Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -0.8416833722      0
#> 2     0.0765382119      0
#> 3    -0.1922890419      0
#> 4     0.3570301790      0
#> 5    -1.0747851799      0
#> 6    -0.7412112881      0
#> 7    -1.1228880099      0
#> 8    -0.8640324497      0
#> 9     0.4151569276      0
#> 10   -1.1281922187      0
#> 11   -0.4775184838      0
#> 12   -0.4547828540      0
#> 13    1.0091338791      0
#> 14    0.8396342719      0
#> 15   -1.1193198384      0
#> 16   -0.3068668509      0
#> 17    0.1994008299      0
#> 18    2.3647743516      0
#> 19   -0.4523329259      0
#> 20   -0.9664538927      0
#> 21   -0.1048319138      0
#> 22    0.5510409941      0
#> 23    0.1562973543      0
#> 24   -1.5560664748      0
#> 25    0.4620669578      0
#> 26    3.1118371525      0
#> 27    0.6255521376      0
#> 28    1.5427920308      0
#> 29   -0.9633823185      0
#> 30   -0.4740360710      0
#> 31    0.6635689746      0
#> 32   -0.6558185045      0
#> 33    0.3058048837      0
#> 34   -0.2300694075      0
#> 35   -0.5430605908      0
#> 36    1.8003554495      0
#> 37   -0.4940561235      0
#> 38   -1.6588393613      0
#> 39    0.6044759828      0
#> 40   -2.0862850159      0
#> 41   -1.6124623978      0
#> 42    0.0500710574      0
#> 43   -1.5830687417      0
#> 44    2.6930765854      0
#> 45   -1.1780632944      0
#> 46   -0.7605706115      0
#> 47   -0.4923782928      0
#> 48   -0.4927736131      0
#> 49   -1.2142034082      0
#> 50    1.3178532184      0
#> 51   -0.2297807118      0
#> 52   -0.0461914559      0
#> 53    1.2181408683      0
#> 54   -0.8356613330      0
#> 55   -0.0392245825      0
#> 56   -0.0041015226      0
#> 57   -2.0339241316      0
#> 58   -0.1694682771      0
#> 59   -0.4796797146      0
#> 60    0.1417237725      0
#> 61    1.6946214378      0
#> 62    0.4763721728      0
#> 63    1.1752021638      0
#> 64    1.0749103640      0
#> 65   -0.4241797385      0
#> 66   -1.6279232187      0
#> 67    0.1825888533      0
#> 68    1.5331525901      0
#> 69   -2.1118486089      0
#> 70   -0.8401894685      0
#> 71    1.1345998030      0
#> 72   -0.9218252767      0
#> 73    2.0366095283      0
#> 74    0.6635422724      0
#> 75   -0.6494022004      0
#> 76    0.9641712987      0
#> 77   -0.8755503611      0
#> 78   -0.3349416621      0
#> 79   -0.8551436726      0
#> 80    1.6665174988      0
#> 81    0.6803653874      0
#> 82    0.6535399444      0
#> 83   -1.4489154795      0
#> 84    0.1707131434      0
#> 85   -0.6150010149      0
#> 86    1.1465959487      0
#> 87   -1.1493902382      0
#> 88    0.5815373877      0
#> 89   -0.2036879876      0
#> 90   -0.5086738004      0
#> 91    0.7953198281      0
#> 92   -0.8310482310      0
#> 93    0.1466828879      0
#> 94   -0.2248750106      0
#> 95    1.0320637452      0
#> 96    0.6353220497      0
#> 97   -0.0130435560      0
#> 98    0.0006224796      0
#> 99   -0.8977431005      0
#> 100   0.4838340167      0
#> 101   1.2148066624      0
#> 102   0.1983320444      0
#> 103   0.7156145831      0
#> 104  -0.7657996159      0
#> 105   0.3307051322      0
#> 106  -1.4824811779      0
#> 107   0.4086735795      0
#> 108   0.1492451115      0
#> 109  -0.3556104752      0
#> 110   0.1477563481      0
#> 111   0.0892449840      0
#> 112  -1.4697562020      0
#> 113  -2.0567040155      0
#> 114   0.3330872428      0
#> 115  -1.1831485862      0
#> 116  -0.0821895607      0
#> 117   0.4197473801      0
#> 118   1.9513777478      0
#> 119  -0.3994991123      0
#> 120  -0.3115067415      0
#> 121  -1.1201813283      0
#> 122  -1.7567863550      0
#> 123   0.5041516780      0
#> 124  -0.9445780136      0
#> 125  -0.2207420089      0
#> 126   1.3900852692      0
#> 127   0.3215445985      0
#> 128  -0.1546774041      0
#> 129  -0.8286283793      0
#> 130   1.5700207964      0
#> 131  -0.7557195320      0
#> 132   0.0028332042      0
#> 133   1.2493584364      0
#> 134   0.4263778135      0
#> 135  -1.0352932062      0
#> 136   1.0385292189      0
#> 137   0.1448530796      0
#> 138  -0.4981186687      0
#> 139   0.4415327340      0
#> 140  -0.0309300586      0
#> 141  -0.8220896669      0
#> 142   1.7904332532      0
#> 143   2.1869664728      0
#> 144  -0.0160539369      0
#> 145   0.3394403937      0
#> 146  -1.1283991303      0
#> 147  -0.3541143840      0
#> 148   0.4898250366      0
#> 149   0.7294482606      0
#> 150  -0.4253847414      0
#> 151   0.7451795828      0
#> 152  -0.7951915182      0
#> 153  -0.9553160313      0
#> 154   2.0316969510      0
#> 155   0.8076544223      0
#> 156   0.1157961706      0
#> 157  -0.1641420107      0
#> 158   0.1916256325      0
#> 159   0.2561480517      0
#> 160   0.4477958154      0
#> 161  -0.3509788601      0
#> 162   1.1283764207      0
#> 163   0.0218972604      0
#> 164  -0.9602541667      0
#> 165  -1.6908992950      0
#> 166   0.3995115267      0
#> 167  -0.2747553719      0
#> 168  -0.0572789372      0
#> 169  -0.1263967794      0
#> 170   1.4050232190      0
#> 171  -1.9017109279      0
#> 172   0.3931958478      0
#> 173  -1.3094124195      0
#> 174   0.0738793078      0
#> 175  -0.2500852757      0
#> 176  -0.1144515507      0
#> 177  -0.1439140775      0
#> 178   1.1384066852      0
#> 179   0.2316475456      0
#> 180   1.1690451442      0
#> 181   0.0625265365      0
#> 182  -0.4065462043      0
#> 183   1.0159948581      0
#> 184   0.7474114369      0
#> 185  -0.7850083896      0
#> 186   0.1845073953      0
#> 187   1.7173692792      0
#> 188   0.8515447930      0
#> 189   0.2675517120      0
#> 190  -0.3994700955      0
#> 191   1.1174693725      0
#> 192   0.3362514570      0
#> 193  -1.4818813611      0
#> 194   1.4037542580      0
#> 195  -1.4047437829      0
#> 196  -0.1100859086      0
#> 197  -0.1475334338      0
#> 198   1.2061070925      0
#> 199  -0.2820509524      0
#> 200   0.2037904157      0
#> 201   0.9965071848      0
#> 202  -0.2855892391      0
#> 203   1.3714288595      0
#> 204  -1.1578077587      0
#> 205  -0.8579588691      0
#> 206   0.3934328150      0
#> 207  -0.8724864742      0
#> 208   1.9106342042      0
#> 209   0.1364502649      0
#> 210   0.5278887931      0
#> 211   0.8709409350      0
#> 212   0.4474594045      0
#> 213   0.2161172029      0
#> 214  -0.7271276819      0
#> 215  -1.4686036120      0
#> 216   0.3472091862      0
#> 217  -1.3529709493      0
#> 218   0.8095284529      0
#> 219   0.5344244118      0
#> 220  -0.5957276914      0
#> 221   0.9621444181      0
#> 222   1.5479571923      0
#> 223   0.3548663536      0
#> 224   0.4746329422      0
#> 225   1.7016369245      0
#> 226   0.4539550157      0
#> 227  -0.1039652017      0
#> 228   2.0415094884      0
#> 229   2.1250541178      0
#> 230   2.7370508996      0
#> 231  -1.1439219083      0
#> 232  -1.8156615542      0
#> 233   0.0181430285      0
#> 234  -1.5575135668      0
#> 235  -1.0265730970      0
#> 236  -0.2326592351      0
#> 237  -1.0324870806      0
#> 238  -0.6477605105      0
#> 239   0.4240677019      0
#> 240   0.4244762813      0
#> 241  -0.1548419648      0
#> 242   0.5946900037      0
#> 243   0.1258087410      0
#> 244   1.7282216418      0
#> 245   1.3673345052      0
#> 246  -0.3117676149      0
#> 247   0.7733115391      0
#> 248   0.5689273301      0
#> 249  -0.8332652472      0
#> 250  -0.3320099864      0
#> 251   1.6262517971      0
#> 252   0.2812635699      0
#> 253  -1.2288729502      0
#> 254   0.8463022978      0
#> 255   2.8209305040      0
#> 256  -0.0945608291      0
#> 257   0.8573396417      0
#> 258   1.7811696984      0
#> 259   1.2372602321      0
#> 260  -1.4521444062      0
#> 261  -0.7699061749      0
#> 262   0.2470360035      0
#> 263  -0.6525828370      0
#> 264  -0.2336879584      0
#> 265  -0.3507186220      0
#> 266  -0.5011545481      0
#> 267   2.0910138777      0
#> 268   1.1398463148      0
#> 269  -1.2756842703      0
#> 270  -1.2656354685      0
#> 271  -1.0469975430      0
#> 272   1.7798804007      0
#> 273   0.6014541104      0
#> 274  -1.3309040657      0
#> 275  -0.8080809948      0
#> 276  -0.0632681424      0
#> 277   0.1900050337      0
#> 278  -0.3781767572      0
#> 279  -0.0336333257      0
#> 280   0.0387423015      0
#> 281   0.2922569264      0
#> 282  -0.8442079593      0
#> 283  -2.1634334673      0
#> 284  -0.2476493726      0
#> 285   0.4835535224      0
#> 286  -0.2219162750      0
#> 287   0.1365521764      0
#> 288   1.1905080552      0
#> 289  -0.0341987145      0
#> 290  -0.2008197274      0
#> 291   0.0975502226      0
#> 292   0.1822841676      0
#> 293  -1.2393056370      0
#> 294   2.0662124238      0
#> 295   1.4974332189      0
#> 296  -0.2841158575      0
#> 297  -0.6223790515      0
#> 298   0.7255477681      0
#> 299   0.3333255691      0
#> 300  -2.8525716833      0
#> 301  -0.7128872222      0
#> 302  -1.3898607741      0
#> 303  -0.2889896141      0
#> 304  -0.7807270692      0
#> 305  -1.0507484726      0
#> 306  -0.4436703206      0
#> 307   1.2739702621      0
#> 308  -0.6695932001      0
#> 309   0.0708335634      0
#> 310  -0.1766272995      0
#> 311  -0.3078194250      0
#> 312   1.8983412991      0
#> 313  -1.0424729977      0
#> 314   0.1223575926      0
#> 315  -0.5356073298      0
#> 316   0.5128438784      0
#> 317  -1.2106530588      0
#> 318   1.2442165477      0
#> 319   0.4858381455      0
#> 320   0.4778668179      0
#> 321  -2.0386188516      0
#> 322  -0.5487485788      0
#> 323  -1.1582585046      0
#> 324   0.0575438035      0
#> 325   1.7514471568      0
#> 326   1.3902142156      0
#> 327  -1.1967324721      0
#> 328  -0.3193328662      0
#> 329  -1.7766007619      0
#> 330  -0.3634950338      0
#> 331  -0.0911697279      0
#> 332  -0.5470295289      0
#> 333  -0.6327000136      0
#> 334   0.9958032001      0
#> 335   0.3295633813      0
#> 336   0.1712965615      0
#> 337  -0.3205199785      0
#> 338  -0.7586081892      0
#> 339  -1.0667306692      0
#> 340   0.5237345247      0
#> 341   0.6720357790      0
#> 342  -1.0460636433      0
#> 343   0.8732101974      0
#> 344  -0.0230134691      0
#> 345   0.1831353948      0
#> 346  -1.3905592166      0
#> 347  -0.9629993310      0
#> 348   0.6311990898      0
#> 349  -1.5563941030      0
#> 350   1.6211780432      0
#> 351   0.2106213176      0
#> 352  -0.3254253981      0
#> 353   1.7925564621      0
#> 354   2.1381750615      0
#> 355  -2.3565423475      0
#> 356   1.6856242568      0
#> 357   1.3108727650      0
#> 358  -0.7690962081      0
#> 359   0.7959718005      0
#> 360   1.8502010478      0
#> 361   0.3152199691      0
#> 362  -1.1449415220      0
#> 363   0.4591209085      0
#> 364   1.8139434184      0
#> 365  -0.3535737418      0
#> 366  -1.1827708769      0
#> 367   0.5354888616      0
#> 368  -0.6373033203      0
#> 369   0.8299979977      0
#> 370  -0.8050619138      0
#> 371   1.7674509745      0
#> 372   1.4700138438      0
#> 373   1.3224262796      0
#> 374   0.2700850872      0
#> 375  -1.1630268432      0
#> 376   0.9628657854      0
#> 377  -0.8623525066      0
#> 378   0.4734754271      0
#> 379   1.9532109106      0
#> 380  -1.8357143641      0
#> 381  -1.2499729655      0
#> 382  -0.4284498199      0
#> 383   0.1047988277      0
#> 384   1.5713009876      0
#> 385  -0.7807094925      0
#> 386  -0.4334823204      0
#> 387   0.0916132733      0
#> 388  -0.0997271611      0
#> 389  -0.2821363627      0
#> 390  -0.0786729580      0
#> 391  -0.3569850431      0
#> 392  -0.0188892264      0
#> 393  -1.5249551077      0
#> 394  -0.2780751999      0
#> 395   0.3553611866      0
#> 396  -0.3718601830      0
#> 397  -1.6583658267      0
#> 398  -0.4346183828      0
#> 399  -0.3768389809      0
#> 400  -0.5013888141      0
#> 401  -0.6860445098      0
#> 402   0.6950020559      0
#> 403  -0.5748206311      0
#> 404   1.4254411373      0
#> 405  -0.6668385289      0
#> 406   0.0863089260      0
#> 407   1.4303700697      0
#> 408  -1.5824779262      0
#> 409  -0.5192451483      0
#> 410   0.8122996219      0
#> 411   1.5798225192      0
#> 412   1.2821830094      0
#> 413  -1.4333516189      0
#> 414  -2.0991702058      0
#> 415  -1.2829469745      0
#> 416  -0.9373716122      0
#> 417   1.4048137064      0
#> 418   0.0727622756      0
#> 419   1.3492347953      0
#> 420  -0.1260798371      0
#> 421   1.4630465726      0
#> 422  -0.6607847235      0
#> 423   0.4551219122      0
#> 424   0.8574914329      0
#> 425  -0.1337573601      0
#> 426   0.1685036910      0
#> 427  -1.7830886024      0
#> 428  -0.4451098819      0
#> 429  -0.8218548591      0
#> 430  -1.1271406432      0
#> 431  -1.2396880182      0
#> 432  -0.8499418624      0
#> 433  -0.9569299035      0
#> 434   1.6487128515      0
#> 435  -0.1353784074      0
#> 436   0.0979378665      0
#> 437   0.7873469743      0
#> 438   0.8930304874      0
#> 439  -0.0787976227      0
#> 440  -0.6988224464      0
#> 441  -0.1948848216      0
#> 442   1.2725757676      0
#> 443   1.2870359563      0
#> 444  -0.5563930580      0
#> 445  -1.1412836079      0
#> 446  -1.5076606590      0
#> 447  -1.3143626217      0
#> 448   1.6502293046      0
#> 449   0.3329878842      0
#> 450  -0.6497406392      0
#> 451   0.1645248651      0
#> 452  -0.6230966662      0
#> 453   0.2087348579      0
#> 454   1.2509343929      0
#> 455   0.6538731278      0
#> 456  -0.7505746856      0
#> 457  -1.4509968591      0
#> 458  -1.8030752412      0
#> 459   2.2282109383      0
#> 460  -1.2961896670      0
#> 461   0.8465807840      0
#> 462   0.3429951272      0
#> 463   1.0192791755      0
#> 464  -0.8132049311      0
#> 465   0.4870501235      0
#> 466  -1.2739507129      0
#> 467   0.6692887867      0
#> 468   1.1378388102      0
#> 469   0.3668968120      0
#> 470  -0.1148049643      0
#> 471   0.5481150601      0
#> 472   0.0646952800      0
#> 473   0.3444236119      0
#> 474  -1.2574635057      0
#> 475   0.4163929437      0
#> 476   1.3366865299      0
#> 477   0.4017611126      0
#> 478  -0.0683445356      0
#> 479  -2.5998609095      0
#> 480   0.5305092317      0
#> 481  -0.0405697178      0
#> 482  -0.9143582432      0
#> 483   0.9218859389      0
#> 484   0.2970211727      0
#> 485   1.6285521599      0
#> 486  -0.2629620987      0
#> 487  -1.4221521870      0
#> 488  -0.0567092464      0
#> 489  -0.0144662520      0
#> 490   0.0878060660      0
#> 491   0.3393820537      0
#> 492  -0.7574756626      0
#> 493  -0.8200236251      0
#> 494   0.6339033138      0
#> 495   0.6246087381      0
#> 496  -0.5308649578      0
#> 497  -1.9903631932      0
#> 498  -3.0339120915      0
#> 499  -1.8394893966      0
#> 500  -1.7355647862      0
#> 501   1.8439873221      0
#> 502  -1.6152079610      0
#> 503   0.0904692540      0
#> 504   0.0518564127      0
#> 505  -1.4115945392      0
#> 506  -1.5764228738      0
#> 507  -0.4299662210      0
#> 508   1.2582690426      0
#> 509  -1.6377625782      0
#> 510  -0.6698624406      0
#> 511   0.8895884941      0
#> 512   0.1088618765      0
#> 513   0.7907881710      0
#> 514   1.2414959920      0
#> 515  -0.7762239239      0
#> 516  -0.3640281721      0
#> 517  -0.0546000968      0
#> 518  -1.7659752728      0
#> 519   0.7155985252      0
#> 520   1.0941423513      0
#> 521  -0.4465860076      0
#> 522  -1.3750420018      0
#> 523  -0.9597193702      0
#> 524   0.0108455815      0
#> 525   0.4725255051      0
#> 526  -0.2793421711      0
#> 527  -0.7586400763      0
#> 528  -0.4751209440      0
#> 529  -1.5967370112      0
#> 530  -1.0616053966      0
#> 531  -0.3152023793      0
#> 532   0.3072010367      0
#> 533  -0.5044160805      0
#> 534   0.1126818448      0
#> 535  -1.5755886706      0
#> 536   0.0193930736      0
#> 537   0.3097749465      0
#> 538   1.1240295635      0
#> 539  -0.2201325501      0
#> 540  -1.5206706120      0
#> 541  -0.3742700126      0
#> 542  -0.0610883761      0
#> 543  -2.4746509033      0
#> 544   0.6960160530      0
#> 545   1.9981648591      0
#> 546   0.1807457131      0
#> 547  -0.6086479719      0
#> 548  -0.0573769115      0
#> 549  -2.1615532575      0
#> 550   0.2513013808      0
#> 551   2.7414914188      0
#> 552  -0.8926270013      0
#> 553  -0.1259003112      0
#> 554  -0.4010679721      0
#> 555  -0.2575379195      0
#> 556   0.5789180284      0
#> 557   0.3133000327      0
#> 558  -0.1799171990      0
#> 559   1.9545548628      0
#> 560   0.4837230984      0
#> 561  -0.4465386138      0
#> 562   1.4047396702      0
#> 563   0.8244849475      0
#> 564   0.2785763060      0
#> 565   1.3594893608      0
#> 566   0.9842191232      0
#> 567  -1.4429073171      0
#> 568  -0.5250589993      0
#> 569  -0.5500274146      0
#> 570   0.1119506334      0
#> 571   0.5546675671      0
#> 572  -0.6666692949      0
#> 573   2.0089762169      0
#> 574   0.1455022547      0
#> 575   1.1856477054      0
#> 576  -0.1817828091      0
#> 577  -0.0824299166      0
#> 578  -0.7275536320      0
#> 579  -0.2489193304      0
#> 580   2.5377403490      0
#> 581  -1.1652586584      0
#> 582   1.5115548286      0
#> 583   0.9656842313      0
#> 584  -0.8139938888      0
#> 585   0.7201604928      0
#> 586  -0.6467741840      0
#> 587  -0.6186879572      0
#> 588   0.0325753994      0
#> 589   0.9880406339      0
#> 590  -0.3608909512      0
#> 591   0.1236454495      0
#> 592  -0.0989624041      0
#> 593  -0.4897128341      0
#> 594   0.5836236337      0
#> 595   0.2634574246      0
#> 596   0.3260980376      0
#> 597   1.0174849593      0
#> 598   0.2117383310      0
#> 599   0.6273602928      0
#> 600   1.5405125571      0
#> 601   0.1221543382      0
#> 602   0.4298675576      0
#> 603   1.1008371533      0
#> 604  -0.6028712084      0
#> 605  -1.4076453187      0
#> 606  -0.5433314016      0
#> 607  -0.0628130612      0
#> 608  -1.3628080531      0
#> 609   1.0999956663      0
#> 610  -0.7525998523      0
#> 611   0.6303220327      0
#> 612  -1.7067279522      0
#> 613   1.4364158326      0
#> 614   1.5626367894      0
#> 615   0.5924305689      0
#> 616  -0.1149847271      0
#> 617   1.2824695856      0
#> 618  -2.9708542163      0
#> 619   1.3678902524      0
#> 620  -1.3694081716      0
#> 621  -1.6323988121      0
#> 622   0.5033750501      0
#> 623  -1.2324024068      0
#> 624   0.7405090279      0
#> 625   2.5242641002      0
#> 626  -1.9428619507      0
#> 627   0.7848892143      0
#> 628  -0.3245280668      0
#> 629   0.0733287682      0
#> 630  -1.6557618804      0
#> 631  -1.3439775175      0
#> 632  -1.3130107481      0
#> 633  -0.2151847792      0
#> 634   0.4259000775      0
#> 635   0.3755182561      0
#> 636  -2.1122054444      0
#> 637   0.2376402057      0
#> 638  -0.7225436632      0
#> 639   0.3910457908      0
#> 640   0.6932971197      0
#> 641   1.1238073003      0
#> 642  -1.1598573521      0
#> 643   0.6688020335      0
#> 644   0.4097378234      0
#> 645  -1.6647197521      0
#> 646   0.4077663805      0
#> 647   0.3169693195      0
#> 648   0.9474908251      0
#> 649  -0.1713649204      0
#> 650  -0.6404502365      0
#> 651  -1.0002292623      0
#> 652  -0.7936800766      0
#> 653  -0.8545130611      0
#> 654   1.2957095146      0
#> 655  -1.6995440097      0
#> 656   0.7945419159      0
#> 657  -0.2597426653      0
#> 658  -0.4064645097      0
#> 659   1.3498985362      0
#> 660   1.4100827743      0
#> 661  -0.9741237978      0
#> 662  -1.3324278398      0
#> 663  -0.5605155593      0
#> 664   0.3785296500      0
#> 665  -1.1288166250      0
#> 666   0.8979323827      0
#> 667   0.6106675126      0
#> 668   0.1160714880      0
#> 669   0.3185800906      0
#> 670   0.3490403294      0
#> 671  -0.8594607678      0
#> 672  -0.1093651379      0
#> 673   0.6139582560      0
#> 674   1.6163664868      0
#> 675  -1.0747682416      0
#> 676   1.3985205320      0
#> 677   0.5400673865      0
#> 678  -0.2096548855      0
#> 679  -0.6930367638      0
#> 680  -0.6418206063      0
#> 681  -0.9230088582      0
#> 682   1.0293734588      0
#> 683   0.3356184769      0
#> 684   1.3418140676      0
#> 685  -0.4557213675      0
#> 686  -1.5913405314      0
#> 687   0.4612887570      0
#> 688   0.8208151362      0
#> 689  -0.1918174593      0
#> 690   0.3346630042      0
#> 691   0.6326239379      0
#> 692   0.8546872057      0
#> 693   0.9562317519      0
#> 694   0.4889837565      0
#> 695  -1.5823353951      0
#> 696  -0.2127617789      0
#> 697  -0.2525761449      0
#> 698  -2.0029860409      0
#> 699  -1.6404516901      0
#> 700   0.7103734623      0
#> 701   0.6227269741      0
#> 702   0.0936322047      0
#> 703  -1.0237974123      0
#> 704  -0.9447164730      0
#> 705   0.5991095776      0
#> 706   0.9307958243      0
#> 707  -0.3941694138      0
#> 708   1.5766069250      0
#> 709   0.4913394656      0
#> 710  -0.4757466532      0
#> 711  -0.5591523065      0
#> 712   0.1680204836      0
#> 713  -0.1365030417      0
#> 714   0.4522936526      0
#> 715  -0.9068558427      0
#> 716   0.3152262197      0
#> 717  -1.2936484347      0
#> 718  -0.5739177138      0
#> 719   1.2744159092      0
#> 720   1.6128348892      0
#> 721  -0.6116985907      0
#> 722   0.5583194600      0
#> 723  -1.1857247962      0
#> 724  -0.0831889926      0
#> 725   0.1698610104      0
#> 726  -0.6237934033      0
#> 727   0.1166383407      0
#> 728  -1.7325266304      0
#> 729   0.0625001226      0
#> 730  -0.8365318362      0
#> 731  -0.7075503356      0
#> 732   0.3861321693      0
#> 733   0.6067340666      0
#> 734   1.1654206355      0
#> 735  -0.5822205035      0
#> 736   1.5023592014      0
#> 737  -0.7798886643      0
#> 738   1.2736653690      0
#> 739   0.0797633779      0
#> 740  -0.8885934827      0
#> 741  -1.2686726615      0
#> 742   1.4844490354      0
#> 743   0.9862851865      0
#> 744  -0.9229647260      0
#> 745   0.6161298551      0
#> 746  -0.7066729434      0
#> 747  -0.9064812569      0
#> 748   1.0342401710      0
#> 749   0.3074602942      0
#> 750   0.8234026249      0
#> 751  -0.9964214367      0
#> 752   0.7277657674      0
#> 753  -0.6949719253      0
#> 754  -0.8534982063      0
#> 755   0.8519211190      0
#> 756  -2.7290864775      0
#> 757  -0.8189481527      0
#> 758  -2.4860145393      0
#> 759  -1.1685923328      0
#> 760   1.0028501681      0
#> 761  -1.2868987966      0
#> 762   0.7022205100      0
#> 763   0.4074110643      0
#> 764   0.2472056044      0
#> 765   1.3756564898      0
#> 766   0.1107149732      0
#> 767   0.4131141766      0
#> 768   1.9045666635      0
#> 769  -0.4316051978      0
#> 770  -0.4307973252      0
#> 771   0.0650023819      0
#> 772   0.3127342209      0
#> 773   0.1918812520      0
#> 774   0.4082267479      0
#> 775   2.5834505580      0
#> 776   0.9367014316      0
#> 777   0.5768916166      0
#> 778   0.6899709640      0
#> 779   0.6425415313      0
#> 780  -0.4713122269      0
#> 781  -1.8664348090      0
#> 782  -1.0759768165      0
#> 783  -0.8894685363      0
#> 784   0.6747071586      0
#> 785   0.0840174274      0
#> 786   0.8785124675      0
#> 787  -0.0763237269      0
#> 788  -0.5401898073      0
#> 789   0.7392134869      0
#> 790  -0.1096778434      0
#> 791   0.8014981713      0
#> 792  -0.3195541350      0
#> 793  -0.3373807340      0
#> 794   0.5698262965      0
#> 795  -1.6448581652      0
#> 796  -0.1808179793      0
#> 797  -0.6310128744      0
#> 798   0.9882892441      0
#> 799   1.7534701812      0
#> 800  -0.3883385860      0
#> 801  -0.2627837039      0
#> 802  -1.0204843386      0
#> 803   0.2459467690      0
#> 804   0.4148821782      0
#> 805  -0.8646413890      0
#> 806  -0.1869711718      0
#> 807   0.0343455114      0
#> 808   1.1135978826      0
#> 809  -0.1244270208      0
#> 810   0.9818657711      0
#> 811   1.4856717540      0
#> 812  -1.0536274064      0
#> 813   0.2471168511      0
#> 814   0.4145801460      0
#> 815   1.2540795593      0
#> 816  -1.4131702186      0
#> 817  -0.7058293269      0
#> 818   1.2250955489      0
#> 819   0.3700960022      0
#> 820   1.1100238085      0
#> 821   0.3719141191      0
#> 822   0.5271511798      0
#> 823   0.1019322726      0
#> 824   1.2547086999      0
#> 825   0.5321376606      0
#> 826   0.2909784743      0
#> 827  -0.4909831518      0
#> 828  -0.3569545673      0
#> 829   0.5064463314      0
#> 830  -2.9103991444      0
#> 831  -0.2012908934      0
#> 832  -0.5592956897      0
#> 833   0.6416364644      0
#> 834  -0.9768612241      0
#> 835  -0.6714532925      0
#> 836   0.1020140249      0
#> 837  -0.7900854629      0
#> 838  -0.4208877291      0
#> 839  -1.6598968278      0
#> 840  -2.2258820475      0
#> 841  -0.5921079881      0
#> 842   1.6149942912      0
#> 843   0.4921787844      0
#> 844   0.5117410889      0
#> 845  -1.7948817271      0
#> 846   0.6719935961      0
#> 847   0.6860149875      0
#> 848   0.6359064468      0
#> 849   2.5454739974      0
#> 850  -0.9337493976      0
#> 851   0.0876328399      0
#> 852   1.3045592662      0
#> 853  -0.3392027353      0
#> 854  -0.5496145366      0
#> 855  -0.0779828789      0
#> 856  -0.1808087628      0
#> 857  -0.2554828226      0
#> 858  -0.2704744235      0
#> 859  -0.3872961934      0
#> 860   0.1280059675      0
#> 861  -1.0157804258      0
#> 862   0.9566584797      0
#> 863   1.0536341959      0
#> 864  -1.1943791215      0
#> 865   0.0816295450      0
#> 866  -0.2909630375      0
#> 867  -1.3512535712      0
#> 868  -0.6790258504      0
#> 869  -0.9024517901      0
#> 870   0.3978855399      0
#> 871   0.2923486754      0
#> 872  -0.2800766013      0
#> 873  -0.0908338264      0
#> 874  -0.7786115277      0
#> 875   0.2298114866      0
#> 876   0.4638877509      0
#> 877  -0.5339034459      0
#> 878   0.6641571261      0
#> 879   0.5681128193      0
#> 880   1.2387360040      0
#> 881   2.0675682298      0
#> 882   0.7843471362      0
#> 883   1.5100601077      0
#> 884  -0.5261757154      0
#> 885   0.6689866555      0
#> 886  -1.5061889995      0
#> 887   1.3148547965      0
#> 888   0.8615596728      0
#> 889  -0.6617842894      0
#> 890  -1.3170726474      0
#> 891  -0.9918651944      0
#> 892   0.7216815493      0
#> 893  -0.3596765521      0
#> 894  -1.4740269546      0
#> 895   0.8954522845      0
#> 896   0.8171406103      0
#> 897   0.3986044329      0
#> 898   1.8133796077      0
#> 899   1.4228859042      0
#> 900   0.1332530734      0
#> 901   0.4708182862      0
#> 902  -0.4833656148      0
#> 903  -0.2743074948      0
#> 904  -0.1694539132      0
#> 905  -1.4462392241      0
#> 906   1.2376065728      0
#> 907   0.4050347605      0
#> 908   1.8406271719      0
#> 909   0.3696368689      0
#> 910   1.4983815115      1
#> 911  -0.7885418527      1
#> 912   2.1424708159      1
#> 913  -0.8619962815      1
#> 914  -0.5216211993      1
#> 915   1.6066900946      1
#> 916   1.3914978107      1
#> 917   0.9974086735      1
#> 918  -1.7145409262      1
#> 919   0.0087130909      1
#> 920  -0.3223489674      1
#> 921   2.7463545327      1
#> 922   0.3365060522      1
#> 923   0.1823676369      1
#> 924  -0.4815444315      1
#> 925   1.2860730017      1
#> 926   0.3377685250      1
#> 927  -0.2601340015      1
#> 928   0.7354355162      1
#> 929   0.0308021226      1
#> 930   0.2971767421      1
#> 931  -0.3685228954      1
#> 932   1.6202448495      1
#> 933  -0.6207593094      1
#> 934   1.9038131444      1
#> 935  -0.2819345675      1
#> 936  -0.2041010179      1
#> 937   2.3351887767      1
#> 938  -0.4846623923      1
#> 939   0.4559710719      1
#> 940   1.3645901399      1
#> 941   1.7853133232      1
#> 942  -0.5356294846      1
#> 943   1.7451443725      1
#> 944  -0.0843172717      1
#> 945   0.8965041282      1
#> 946   0.9716644753      1
#> 947   1.1879677035      1
#> 948  -0.2451679104      1
#> 949   0.1347310608      1
#> 950   1.4544024752      1
#> 951   0.5006162578      1
#> 952  -0.0013507096      1
#> 953   0.3588914266      1
#> 954   0.6098529089      1
#> 955   1.2229716100      1
#> 956   1.5881460329      1
#> 957  -0.4730208667      1
#> 958   0.9376541561      1
#> 959   2.0209076307      1
#> 960   1.3652783106      1
#> 961   1.2628355258      1
#> 962   1.4592736318      1
#> 963   0.5079260900      1
#> 964  -0.1775262676      1
#> 965   0.3680194398      1
#> 966   2.2077370098      1
#> 967   1.8769036195      1
#> 968  -1.2070329169      1
#> 969   0.9153621987      1
#> 970   0.7588017952      1
#> 971   2.0649411737      1
#> 972   1.3235934232      1
#> 973   3.2899528876      1
#> 974  -0.0777758389      1
#> 975   0.6816729492      1
#> 976  -0.2632059988      1
#> 977  -0.4132227257      1
#> 978  -0.0958678052      1
#> 979   1.3885878761      1
#> 980  -0.5623881793      1
#> 981   0.7370485268      1
#> 982   0.3888441828      1
#> 983   0.9473496193      1
#> 984   2.0268700552      1
#> 985   1.5893305166      1
#> 986  -0.3382718992      1
#> 987   0.4453244866      1
#> 988   1.4853640174      1
#> 989   1.5601846176      1
#> 990   0.9475671249      1
#> 991  -0.7986489145      1
#> 992  -0.3955501402      1
#> 993   0.8510066956      1
#> 994   1.0374044484      1
#> 995   0.7568749154      1
#> 996  -1.7247400679      1
#> 997   0.6576135759      1
#> 998   2.6596931404      1
#> 999  -0.8076187079      1
#> 1000  0.2896096538      1