Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     1.5995634740      0
#> 2    -1.0504275256      0
#> 3     2.2202370742      0
#> 4     0.2616911295      0
#> 5    -0.4916163566      0
#> 6    -1.0201610283      0
#> 7    -0.3602708245      0
#> 8     1.3346727068      0
#> 9     0.7430841299      0
#> 10   -1.9376646090      0
#> 11    1.5656175394      0
#> 12    0.5285287942      0
#> 13   -1.1516225892      0
#> 14   -0.2724342637      0
#> 15    0.9666308246      0
#> 16   -1.3198658114      0
#> 17   -1.2618491028      0
#> 18    0.1370225725      0
#> 19    0.0901952783      0
#> 20    1.6288487908      0
#> 21    0.8213297033      0
#> 22    1.4265073005      0
#> 23    0.8751283064      0
#> 24   -0.8088726091      0
#> 25    0.3045181949      0
#> 26    0.4409806899      0
#> 27   -0.7236023906      0
#> 28   -0.8820565254      0
#> 29   -0.1604907971      0
#> 30   -0.8031669173      0
#> 31   -1.1766983938      0
#> 32   -0.2022626995      0
#> 33    0.0164744523      0
#> 34    0.5547230726      0
#> 35   -1.8877839427      0
#> 36   -1.0351370781      0
#> 37    0.5088132478      0
#> 38   -0.4470482171      0
#> 39   -0.4194145885      0
#> 40    0.2402711399      0
#> 41   -0.1105222554      0
#> 42   -0.7711600032      0
#> 43    0.7078879541      0
#> 44   -0.2727483333      0
#> 45    1.1078318298      0
#> 46    0.0970018601      0
#> 47    1.0442166838      0
#> 48    1.4626310070      0
#> 49   -0.6649596159      0
#> 50   -1.2660605589      0
#> 51   -0.7627093562      0
#> 52   -0.4461008166      0
#> 53   -0.1593304169      0
#> 54    0.5724564975      0
#> 55   -1.1575326255      0
#> 56   -1.1681970849      0
#> 57    0.5617256102      0
#> 58    2.1353159143      0
#> 59   -1.5915674339      0
#> 60   -1.4297721266      0
#> 61    0.4702562950      0
#> 62    0.0959981435      0
#> 63    0.4279210059      0
#> 64    0.6937310335      0
#> 65    0.2949071705      0
#> 66   -1.1725945923      0
#> 67   -0.5300970705      0
#> 68   -1.2362224315      0
#> 69   -0.8862372976      0
#> 70    0.1650527674      0
#> 71    0.4326755097      0
#> 72   -0.6525153882      0
#> 73    0.1760778169      0
#> 74    1.5898926896      0
#> 75   -1.0510501104      0
#> 76   -1.1953857745      0
#> 77   -1.1964044419      0
#> 78    0.1615525873      0
#> 79    0.6018464198      0
#> 80    0.1953965323      0
#> 81    0.4637658685      0
#> 82    0.1780106527      0
#> 83   -0.8505436883      0
#> 84   -0.9088292877      0
#> 85   -0.2799642056      0
#> 86    0.8234078203      0
#> 87   -0.8740010640      0
#> 88   -0.4556972782      0
#> 89   -0.3829792874      0
#> 90   -0.2158676403      0
#> 91    0.7518411719      0
#> 92   -0.0810168351      0
#> 93    0.2473679610      0
#> 94   -0.0457876294      0
#> 95   -0.1218605914      0
#> 96   -0.2236312754      0
#> 97   -1.7393924533      0
#> 98    1.4466092140      0
#> 99   -0.0705183556      0
#> 100   0.8903871959      0
#> 101   0.6042428424      0
#> 102  -0.1091284894      0
#> 103  -0.1827053771      0
#> 104  -2.3222103440      0
#> 105  -0.2399791794      0
#> 106   0.5220314746      0
#> 107  -0.1820455149      0
#> 108  -0.9088829931      0
#> 109   0.1095648947      0
#> 110  -1.7241856033      0
#> 111  -1.3427557421      0
#> 112   0.0376629189      0
#> 113   1.3175848319      0
#> 114  -1.1035191792      0
#> 115   0.0490550928      0
#> 116   0.6035945367      0
#> 117   0.1842087791      0
#> 118  -1.7152699329      0
#> 119  -0.4508713605      0
#> 120  -1.6333332085      0
#> 121   1.4113209494      0
#> 122   0.8114081139      0
#> 123   0.1482588965      0
#> 124  -1.4016998691      0
#> 125   1.5561328397      0
#> 126  -0.4294012412      0
#> 127   0.9945348359      0
#> 128   1.1952728660      0
#> 129  -1.7841131045      0
#> 130   0.6862188775      0
#> 131  -1.7851179963      0
#> 132  -1.4891881413      0
#> 133   2.2756077863      0
#> 134  -0.9524591035      0
#> 135   2.1651487017      0
#> 136   0.0880114168      0
#> 137   0.6466877827      0
#> 138  -0.3426091382      0
#> 139  -0.8510542515      0
#> 140   0.6482992479      0
#> 141  -1.0714172808      0
#> 142  -1.7688941078      0
#> 143   2.3497204639      0
#> 144   0.2369479568      0
#> 145   0.0128191246      0
#> 146   2.0030780090      0
#> 147   0.0470379298      0
#> 148  -2.5073573017      0
#> 149   0.4209575620      0
#> 150   1.1123895809      0
#> 151   0.4494905897      0
#> 152   2.3993817918      0
#> 153  -1.0905119927      0
#> 154   0.6902067919      0
#> 155  -0.2383615230      0
#> 156  -1.0062749907      0
#> 157   0.1805623300      0
#> 158  -0.7246360373      0
#> 159   1.3891965472      0
#> 160   2.3709378886      0
#> 161   0.5107198779      0
#> 162   0.6714838643      0
#> 163   0.4101912935      0
#> 164  -1.2721292654      0
#> 165   0.2933273802      0
#> 166  -0.2006190801      0
#> 167   0.8783109405      0
#> 168   1.3470910399      0
#> 169   0.4050313231      0
#> 170  -0.1405525696      0
#> 171   0.4537083909      0
#> 172   0.4634539731      0
#> 173  -1.2554786895      0
#> 174  -1.0082148288      0
#> 175   1.5523305571      0
#> 176   1.3996424657      0
#> 177   0.4158093600      0
#> 178  -0.6427324660      0
#> 179   0.7217075579      0
#> 180   0.2939209372      0
#> 181  -0.6556599559      0
#> 182  -0.1795003113      0
#> 183   1.5323091949      0
#> 184   0.0702125548      0
#> 185   0.1720594130      0
#> 186  -2.0127821849      0
#> 187   1.6384803653      0
#> 188   1.0629025173      0
#> 189  -1.8246537593      0
#> 190  -0.1393069580      0
#> 191   3.1031817687      0
#> 192  -0.1042079510      0
#> 193   2.4319771207      0
#> 194  -0.0906129728      0
#> 195   0.0526953388      0
#> 196  -2.0424119457      0
#> 197  -1.6460408076      0
#> 198  -0.1764322030      0
#> 199   0.4141645530      0
#> 200  -0.2534382935      0
#> 201   0.9631899935      0
#> 202  -0.8554025185      0
#> 203  -2.6656603090      0
#> 204   0.9292943998      0
#> 205   0.8345075466      0
#> 206  -2.6649378086      0
#> 207   0.1804153565      0
#> 208   2.2204080296      0
#> 209   0.3157124030      0
#> 210  -1.0647370047      0
#> 211  -0.0031648930      0
#> 212   1.2628487546      0
#> 213  -0.8417165758      0
#> 214   0.4208805536      0
#> 215  -0.0206524224      0
#> 216  -1.8670137174      0
#> 217   1.4420301550      0
#> 218  -0.0707461969      0
#> 219  -0.4164530792      0
#> 220  -0.7142754829      0
#> 221   1.3544477357      0
#> 222  -1.0067044856      0
#> 223  -0.2746320797      0
#> 224  -0.4031626102      0
#> 225  -0.3345154548      0
#> 226   1.1010933537      0
#> 227  -0.2234614997      0
#> 228  -0.0471221861      0
#> 229   0.6759039961      0
#> 230  -0.2846242951      0
#> 231   0.3946691983      0
#> 232  -0.7007489244      0
#> 233  -0.5215809035      0
#> 234  -0.9143718631      0
#> 235   0.2427876939      0
#> 236   0.4188528149      0
#> 237  -0.0872239122      0
#> 238   0.9833747132      0
#> 239   1.6321874606      0
#> 240   0.4865378050      0
#> 241   0.9090645474      0
#> 242   0.0213839374      0
#> 243   0.1743775151      0
#> 244  -1.5133644910      0
#> 245  -0.7129102445      0
#> 246   0.2347033660      0
#> 247   1.1355747858      0
#> 248   0.9811061807      0
#> 249   0.4247994601      0
#> 250   0.0699963987      0
#> 251  -1.3042408975      0
#> 252   1.4723734045      0
#> 253   0.6617669280      0
#> 254   2.5805298995      0
#> 255  -0.1657962181      0
#> 256   3.2774084353      0
#> 257   0.0502321308      0
#> 258   0.7416471341      0
#> 259   0.6218123814      0
#> 260   0.2394997998      0
#> 261  -1.4247176457      0
#> 262  -1.4566923250      0
#> 263   0.5600437624      0
#> 264   0.9992751683      0
#> 265   0.5392169554      0
#> 266  -0.7396772094      0
#> 267   0.6191304767      0
#> 268   1.1252820360      0
#> 269   0.3584913291      0
#> 270   0.1909720456      0
#> 271   0.3852038882      0
#> 272  -1.1658353439      0
#> 273   0.3821811109      0
#> 274   0.6688976238      0
#> 275   0.3852543384      0
#> 276   0.4891713230      0
#> 277  -1.1954557742      0
#> 278   0.5308591203      0
#> 279  -0.8197616914      0
#> 280   0.5507551777      0
#> 281  -0.8443648030      0
#> 282   0.7049229736      0
#> 283   0.7267262110      0
#> 284   0.5889545415      0
#> 285   0.7793545590      0
#> 286   0.5341279813      0
#> 287  -0.3460294017      0
#> 288  -0.2932239329      0
#> 289   0.5186525092      0
#> 290   1.2044064292      0
#> 291  -1.4808439710      0
#> 292  -0.0894064487      0
#> 293  -0.4882273531      0
#> 294  -0.1630072962      0
#> 295  -0.2560911315      0
#> 296  -1.2470677453      0
#> 297  -0.4129413842      0
#> 298  -0.2393868941      0
#> 299   0.8531251062      0
#> 300   0.9096919303      0
#> 301   0.8456481431      0
#> 302  -1.1612740998      0
#> 303   1.0376325827      0
#> 304   0.8769640212      0
#> 305  -0.7568565808      0
#> 306   0.1498029741      0
#> 307   0.1385278643      0
#> 308   0.1287896537      0
#> 309   0.8113389404      0
#> 310  -0.7520233311      0
#> 311  -0.1589729731      0
#> 312   0.2018407330      0
#> 313  -0.6721841629      0
#> 314  -0.1765725117      0
#> 315  -0.9855543227      0
#> 316  -1.1248246053      0
#> 317  -1.3989434138      0
#> 318  -0.0379630641      0
#> 319  -0.4193926577      0
#> 320   0.3353577045      0
#> 321  -0.9160230584      0
#> 322   0.9062363387      0
#> 323   1.4080721609      0
#> 324   0.1463220490      0
#> 325   1.1121041459      0
#> 326  -0.5523009459      0
#> 327  -3.1496681870      0
#> 328   0.3697098573      0
#> 329  -1.4479039141      0
#> 330   0.8237419131      0
#> 331   0.2943889629      0
#> 332  -1.6101738933      0
#> 333   0.7842467740      0
#> 334  -0.1829375028      0
#> 335  -0.5145123218      0
#> 336   0.6621635202      0
#> 337  -0.8069996982      0
#> 338  -0.2621441829      0
#> 339  -0.0559171799      0
#> 340  -0.6770636220      0
#> 341  -0.6066725036      0
#> 342   0.0807631697      0
#> 343  -1.2605574057      0
#> 344  -0.1068278679      0
#> 345   0.4623687995      0
#> 346  -0.2149686850      0
#> 347  -0.0043603215      0
#> 348  -0.8306552195      0
#> 349  -0.6816735810      0
#> 350  -0.1823048092      0
#> 351  -0.5301390599      0
#> 352   1.3346040270      0
#> 353  -0.5802161682      0
#> 354  -0.3751233679      0
#> 355  -0.0775087526      0
#> 356   1.2811827592      0
#> 357  -0.0282487076      0
#> 358  -1.3268324913      0
#> 359  -1.2257831165      0
#> 360   0.4994741868      0
#> 361   1.2170568860      0
#> 362  -1.8977306763      0
#> 363   1.2834145319      0
#> 364  -1.1138384195      0
#> 365  -0.3161108676      0
#> 366   0.3847095677      0
#> 367  -2.3246000005      0
#> 368  -1.5157353140      0
#> 369  -1.0058414701      0
#> 370   0.0003528633      0
#> 371   0.2672619994      0
#> 372  -0.0344821988      0
#> 373  -1.0243400261      0
#> 374  -0.1437218150      0
#> 375  -2.7761309300      0
#> 376   1.9454059959      0
#> 377  -0.4042978169      0
#> 378   2.2765193904      0
#> 379  -0.0782543050      0
#> 380  -0.4643842197      0
#> 381   1.2554644383      0
#> 382   0.5824245848      0
#> 383  -1.2374467192      0
#> 384   0.9696315624      0
#> 385  -0.6060142211      0
#> 386   0.3667434492      0
#> 387   1.4232634825      0
#> 388  -0.7260827181      0
#> 389  -0.6483051969      0
#> 390  -1.1601262223      0
#> 391  -1.2717601266      0
#> 392   0.8474765230      0
#> 393   0.1058623212      0
#> 394   1.3574157992      0
#> 395  -0.8318148122      0
#> 396  -0.7137551678      0
#> 397  -0.9370129437      0
#> 398   0.2839506025      0
#> 399   0.0543881863      0
#> 400  -0.6591922921      0
#> 401  -1.4486156561      0
#> 402  -0.9036758845      0
#> 403   0.9502983274      0
#> 404  -0.5727362069      0
#> 405  -1.1769979701      0
#> 406   0.3633753734      0
#> 407   1.2963182462      0
#> 408  -0.4537986490      0
#> 409  -2.0698263643      0
#> 410  -1.2278535035      0
#> 411   0.5058606149      0
#> 412   1.5714040766      0
#> 413  -0.4838269573      0
#> 414  -0.5682731282      0
#> 415  -0.0098380631      0
#> 416   1.8353700686      0
#> 417  -0.5670794940      0
#> 418  -0.7906273806      0
#> 419  -0.7724864944      0
#> 420  -1.4507648104      0
#> 421   1.7991230910      0
#> 422   1.1901582804      0
#> 423  -1.2778831853      0
#> 424   0.4574343335      0
#> 425  -0.8763047622      0
#> 426   1.7994300619      0
#> 427   1.2498602922      0
#> 428   0.9392971215      0
#> 429   0.6675857384      0
#> 430  -0.3220745524      0
#> 431   1.0272864946      0
#> 432   0.6332169463      0
#> 433  -0.8964092102      0
#> 434  -0.8809962918      0
#> 435   0.3847109276      0
#> 436  -1.2364164507      0
#> 437  -1.4948264257      0
#> 438  -0.2576003508      0
#> 439  -1.3536017087      0
#> 440  -1.5297687633      0
#> 441   1.6205767175      0
#> 442   0.2999600200      0
#> 443   0.7123689916      0
#> 444  -0.4217120880      0
#> 445   0.8336803404      0
#> 446  -0.8657391459      0
#> 447  -0.9722503928      0
#> 448  -0.4641310667      0
#> 449  -1.2655364563      0
#> 450   2.9916615318      0
#> 451  -0.1571823182      0
#> 452   0.4725396033      0
#> 453   1.5474280640      0
#> 454   0.5245802235      0
#> 455   0.9805961283      0
#> 456  -0.7258028270      0
#> 457  -0.7513368291      0
#> 458  -0.0239748156      0
#> 459  -0.0997595504      0
#> 460   2.0912701783      0
#> 461   1.1367744722      0
#> 462   0.9622881686      0
#> 463  -1.1221729916      0
#> 464   1.3810299957      0
#> 465  -1.0635624531      0
#> 466  -1.4296046491      0
#> 467   0.6145545071      0
#> 468   0.0543769713      0
#> 469  -0.1112362882      0
#> 470   1.7907014765      0
#> 471   1.0184666447      0
#> 472   0.2538563353      0
#> 473   0.7262263557      0
#> 474   0.4400162190      0
#> 475   1.3565664964      0
#> 476   1.1822837104      0
#> 477   0.0890178987      0
#> 478  -1.1875821349      0
#> 479   0.2781078071      0
#> 480  -0.8417024060      0
#> 481  -0.3859323921      0
#> 482   1.5781240720      0
#> 483   0.4701199297      0
#> 484  -1.1265549806      0
#> 485  -0.3985092658      0
#> 486   0.3361749741      0
#> 487  -1.3293123992      0
#> 488  -0.3170315163      0
#> 489  -1.7744480662      0
#> 490   0.3132601778      0
#> 491   0.0396110878      0
#> 492  -0.4679330976      0
#> 493   1.3547667651      0
#> 494   1.0428889138      0
#> 495   0.2559769227      0
#> 496   1.1889794286      0
#> 497   0.3664872894      0
#> 498   0.2649868083      0
#> 499   1.1023082256      0
#> 500   0.0499933488      0
#> 501  -0.1581868049      0
#> 502  -0.9175192433      0
#> 503   2.4124763522      0
#> 504  -0.6119598653      0
#> 505   0.6290911814      0
#> 506   1.3072412915      0
#> 507  -0.3372584047      0
#> 508  -0.9978192545      0
#> 509  -1.5088564607      0
#> 510  -0.4519936099      0
#> 511   0.9538003564      0
#> 512   0.4488644689      0
#> 513  -0.3482561803      0
#> 514  -0.3001992260      0
#> 515   0.2031148114      0
#> 516   2.0806922049      0
#> 517   1.4579694317      0
#> 518  -0.3601270534      0
#> 519  -0.2817297295      0
#> 520  -0.2147161741      0
#> 521  -1.9389289295      0
#> 522  -0.2402362833      0
#> 523   0.6527372006      0
#> 524  -0.4900865067      0
#> 525   1.3418932836      0
#> 526  -0.9182383957      0
#> 527   0.4275496511      0
#> 528  -1.1249188414      0
#> 529   1.2867044556      0
#> 530  -1.3781731371      0
#> 531   1.2467084931      0
#> 532  -0.9720570631      0
#> 533   1.1011388230      0
#> 534  -1.2569808415      0
#> 535  -2.1159050063      0
#> 536  -0.9523156033      0
#> 537  -1.4178668961      0
#> 538   0.7108517145      0
#> 539   0.4111738632      0
#> 540  -0.7336672171      0
#> 541   0.3047175210      0
#> 542   0.8965628940      0
#> 543   0.7004788060      0
#> 544  -0.4459930354      0
#> 545  -0.1195927144      0
#> 546  -0.3772761132      0
#> 547   0.5002833790      0
#> 548   0.8484033325      0
#> 549  -1.0979976107      0
#> 550   0.4979792864      0
#> 551   0.5780026549      0
#> 552  -0.2320590103      0
#> 553  -0.6788018651      0
#> 554  -0.2726570302      0
#> 555   1.8697401761      0
#> 556  -0.5025119724      0
#> 557   1.1500932106      0
#> 558   0.8264884674      0
#> 559   2.0864840019      0
#> 560   1.0987679314      0
#> 561  -1.0149090478      0
#> 562   0.5536048308      0
#> 563  -1.3678156061      0
#> 564   0.6478251682      0
#> 565  -1.5937582658      0
#> 566  -1.3721313065      0
#> 567  -0.0752910599      0
#> 568   0.3930895817      0
#> 569  -1.0130208304      0
#> 570  -0.4749031399      0
#> 571   1.3073277531      0
#> 572  -1.1429138376      0
#> 573   0.7982141477      0
#> 574   0.1953532242      0
#> 575   0.8287811572      0
#> 576   0.2588510274      0
#> 577   0.1439292331      0
#> 578   1.5127827570      0
#> 579  -1.3312719813      0
#> 580  -1.4893347548      0
#> 581   0.5019318608      0
#> 582   0.4810629103      0
#> 583   0.6186511737      0
#> 584   0.3815767443      0
#> 585   1.0980009794      0
#> 586  -1.9163262428      0
#> 587  -0.8247954500      0
#> 588   0.6332319784      0
#> 589  -0.2042794901      0
#> 590  -0.4208675915      0
#> 591   0.7184750482      0
#> 592   0.6861445424      0
#> 593   0.0629901278      0
#> 594   0.3936776177      0
#> 595  -1.1230634532      0
#> 596   1.4124708402      0
#> 597   0.9324883357      0
#> 598   1.2644095070      0
#> 599  -0.1659300079      0
#> 600  -1.3611199959      0
#> 601  -0.3375176444      0
#> 602   0.2880389478      0
#> 603  -0.0364672496      0
#> 604  -0.6173189518      0
#> 605   0.0769405652      0
#> 606  -0.9489201447      0
#> 607   0.4488585488      0
#> 608   0.6452103107      0
#> 609   2.3489441288      0
#> 610   3.1809340080      0
#> 611  -0.1712698249      0
#> 612   1.3629375396      0
#> 613  -0.2904957526      0
#> 614   0.8036374456      0
#> 615  -0.2683268593      0
#> 616  -0.3717235386      0
#> 617  -0.4155821274      0
#> 618  -1.3886544004      0
#> 619  -0.7969189120      0
#> 620   0.4373662750      0
#> 621   0.4774352824      0
#> 622   1.5203585970      0
#> 623   1.1868129029      0
#> 624  -1.8181846092      0
#> 625  -1.5730646275      0
#> 626   0.7982964142      0
#> 627   0.5447503759      0
#> 628  -0.4546403769      0
#> 629   1.0668795810      0
#> 630  -1.0819911082      0
#> 631  -1.7976588325      0
#> 632   0.4608475193      0
#> 633  -1.8431565340      0
#> 634  -0.3337967755      0
#> 635   0.7779694298      0
#> 636  -1.4310207600      0
#> 637  -0.3648185201      0
#> 638  -1.8521927479      0
#> 639  -0.6997177329      0
#> 640   1.3333980153      0
#> 641  -1.2792939087      0
#> 642   0.8615194578      0
#> 643   0.6812623317      0
#> 644   0.7419862309      0
#> 645   0.1121028671      0
#> 646   0.3035215208      0
#> 647  -1.2752711830      0
#> 648   0.6696477509      0
#> 649  -0.3980555983      0
#> 650   0.0793448349      0
#> 651   1.4982301672      0
#> 652  -0.0895487353      0
#> 653   0.8397984133      0
#> 654  -0.8920331989      0
#> 655   0.2130322045      0
#> 656   0.5276031542      0
#> 657   0.6334545121      0
#> 658   0.5853707766      0
#> 659  -0.4373809130      0
#> 660   1.0355824569      0
#> 661   1.0365444083      0
#> 662  -1.3985687165      0
#> 663  -0.6085861593      0
#> 664   0.8995698145      0
#> 665   1.4410967474      0
#> 666  -0.2164198734      0
#> 667   0.6567816687      0
#> 668   0.9722724186      0
#> 669   0.0571147507      0
#> 670   0.5084260173      0
#> 671   0.8657091641      0
#> 672   1.1062460254      0
#> 673  -0.5360878064      0
#> 674   0.8608278612      0
#> 675   0.2957246388      0
#> 676   0.8992960981      0
#> 677   1.2256390441      0
#> 678   1.9845212939      0
#> 679  -0.6756662413      0
#> 680  -0.6911120318      0
#> 681   1.6068658759      0
#> 682   2.4293903773      0
#> 683  -0.0120282079      0
#> 684  -1.2844150788      0
#> 685   1.2344438853      0
#> 686  -0.0552437028      0
#> 687  -0.6123317932      0
#> 688  -0.0230328076      0
#> 689  -0.2506831432      0
#> 690   0.7669429150      0
#> 691   1.1575127107      0
#> 692  -0.8514880548      0
#> 693  -0.6024253156      0
#> 694  -1.1080683710      0
#> 695  -0.5950540306      0
#> 696   0.2492766145      0
#> 697  -0.5993975858      0
#> 698   2.6579899516      0
#> 699   1.7713521929      0
#> 700   0.6230649501      0
#> 701  -1.0188673074      0
#> 702  -0.5884247504      0
#> 703   0.8075971715      0
#> 704   0.0826456925      0
#> 705   0.7587537683      0
#> 706  -0.8950212006      0
#> 707   0.5209960892      0
#> 708   1.3103625259      0
#> 709   1.1757803851      0
#> 710  -0.9509290913      0
#> 711  -0.4957154310      0
#> 712   0.2522843037      0
#> 713  -0.9802897048      0
#> 714  -0.6020969699      0
#> 715  -0.1402556268      0
#> 716  -0.7392623367      0
#> 717   0.2730215193      0
#> 718   1.6271324209      0
#> 719   0.3783660424      0
#> 720   0.4443379245      0
#> 721   0.5205428759      0
#> 722   0.2206822260      0
#> 723  -0.0032349601      0
#> 724  -0.3797935049      0
#> 725  -0.2948196350      0
#> 726  -1.0703516524      0
#> 727   1.1816267239      0
#> 728   0.3592323391      0
#> 729  -1.0794903232      0
#> 730   0.1880079552      0
#> 731  -1.9936738673      0
#> 732   0.8935333125      0
#> 733  -0.0457021193      0
#> 734   1.3879922988      0
#> 735   0.2968815317      0
#> 736   0.2667148736      0
#> 737  -0.7072628120      0
#> 738  -1.7040297190      0
#> 739   0.0648396643      0
#> 740  -0.7433530444      0
#> 741  -0.1976684947      0
#> 742   0.9483497435      0
#> 743  -0.3408292511      0
#> 744  -0.1154545344      0
#> 745   1.1066955189      0
#> 746  -1.0651436475      0
#> 747   0.8666749425      0
#> 748   0.3948425575      0
#> 749   0.9522415891      0
#> 750  -0.9150384271      0
#> 751  -2.3484247877      0
#> 752  -0.0858868016      0
#> 753  -0.9653632816      0
#> 754   0.0052697116      0
#> 755  -1.5663653447      0
#> 756  -0.9364627421      0
#> 757  -1.1955688988      0
#> 758   1.8149958387      0
#> 759  -1.1098262853      0
#> 760   0.4805328259      0
#> 761   0.7361834815      0
#> 762   0.5331220225      0
#> 763   0.5521388024      0
#> 764   1.5294876362      0
#> 765  -0.4165226337      0
#> 766   0.6372259043      0
#> 767   0.4185545885      0
#> 768  -1.5918146251      0
#> 769  -0.7838109692      0
#> 770  -0.6769121967      0
#> 771  -0.2484122818      0
#> 772   0.4111114932      0
#> 773  -0.8490108168      0
#> 774   0.5731980944      0
#> 775   1.6484348147      0
#> 776  -0.4961423778      0
#> 777   1.1839393914      0
#> 778   0.6645059285      0
#> 779   1.3749996453      0
#> 780  -2.3421364822      0
#> 781   0.5353939378      0
#> 782   0.0123326498      0
#> 783  -1.5019721994      0
#> 784  -0.5952082300      0
#> 785   2.1005241334      0
#> 786  -0.6023426627      0
#> 787   0.0359748464      0
#> 788  -0.1515354718      0
#> 789  -0.2216929617      0
#> 790  -0.0588187776      0
#> 791  -1.6676415405      0
#> 792  -0.7628595595      0
#> 793   0.0021018239      0
#> 794  -0.4094559958      0
#> 795   0.0361268300      0
#> 796   0.2511582008      0
#> 797   0.5356680931      0
#> 798   1.3578655340      0
#> 799   1.1715668358      0
#> 800  -2.1758715289      0
#> 801   0.2969489596      0
#> 802   1.3465786729      0
#> 803  -0.4482581522      0
#> 804  -0.1491245990      0
#> 805   0.6672666941      0
#> 806   0.7306930946      0
#> 807  -0.7291328465      0
#> 808  -0.1807045895      0
#> 809   0.0378423905      0
#> 810   1.3666594379      0
#> 811   0.0618047917      0
#> 812  -0.3862443110      0
#> 813   0.5141964895      0
#> 814   1.3338236763      0
#> 815  -0.3150652994      0
#> 816  -0.9552495049      0
#> 817   0.0958989417      0
#> 818   1.0616245703      0
#> 819  -0.8465513391      0
#> 820  -1.3839466533      0
#> 821   1.8441094649      0
#> 822  -2.1762691705      0
#> 823  -0.2184627635      0
#> 824   0.4582615017      0
#> 825   2.0289273417      0
#> 826  -0.3492979244      0
#> 827   1.7320591516      0
#> 828   0.6800029775      0
#> 829   1.8937943704      0
#> 830   0.1009852294      0
#> 831  -0.3186735908      0
#> 832   1.4935538975      0
#> 833   0.6561848773      0
#> 834  -1.6659842678      0
#> 835  -0.4331516392      0
#> 836   0.1620608790      0
#> 837  -0.6425332701      0
#> 838  -0.0389302965      0
#> 839   1.3590513009      0
#> 840   0.6646658348      0
#> 841  -0.7044921486      0
#> 842  -0.9927222779      0
#> 843   0.1556571319      0
#> 844   1.2023977232      0
#> 845   1.2729074850      0
#> 846   0.3577901638      0
#> 847   0.1830627421      0
#> 848   0.1517440152      0
#> 849   1.2605087582      0
#> 850  -1.1958688877      0
#> 851  -0.9274457274      0
#> 852   1.9459166714      0
#> 853   0.0817717120      0
#> 854   0.2166511658      0
#> 855   0.9740863146      0
#> 856   0.1170188373      0
#> 857  -0.4734869648      0
#> 858   0.5681082378      0
#> 859   0.2549348508      0
#> 860  -0.6054243029      0
#> 861  -1.4820893157      0
#> 862   0.6412828014      0
#> 863   2.1542888379      0
#> 864  -1.5162978466      0
#> 865  -0.5708709613      0
#> 866  -2.7383523078      0
#> 867  -0.2692396537      0
#> 868  -0.5444837296      0
#> 869   0.3458975213      0
#> 870   0.0503658515      0
#> 871  -0.0909793130      0
#> 872  -1.7943415364      0
#> 873   1.1935787062      0
#> 874   0.7562543289      0
#> 875  -1.0331921357      0
#> 876   0.8183749377      0
#> 877   0.0598127350      0
#> 878   0.2005388996      0
#> 879   1.3123222306      0
#> 880   1.0225029791      0
#> 881  -1.3922525102      0
#> 882   1.0044692297      0
#> 883   0.8399993664      0
#> 884  -0.1938607553      0
#> 885   0.0266953591      0
#> 886  -0.1066578817      1
#> 887  -0.0204416171      1
#> 888  -0.7409482541      1
#> 889  -0.8262953009      1
#> 890   1.6670823705      1
#> 891   0.6279027121      1
#> 892  -0.0235231115      1
#> 893   2.0507443927      1
#> 894   1.1709672678      1
#> 895   0.8524320906      1
#> 896   1.5625809480      1
#> 897  -0.2345983293      1
#> 898   0.5474174835      1
#> 899   1.1033344267      1
#> 900   0.2036879777      1
#> 901   0.8230125403      1
#> 902  -0.3693364498      1
#> 903   1.8540871905      1
#> 904   2.2777624541      1
#> 905   0.0520326323      1
#> 906   2.0498370813      1
#> 907   0.4015575190      1
#> 908   0.7349098372      1
#> 909   0.8926330509      1
#> 910   1.1102463344      1
#> 911   0.6970230633      1
#> 912   0.7333264385      1
#> 913   1.1946424911      1
#> 914   0.3964481260      1
#> 915  -0.1400819008      1
#> 916   0.0884792486      1
#> 917   0.0549107114      1
#> 918   0.8948364753      1
#> 919   0.2446216853      1
#> 920  -1.2017657439      1
#> 921   1.0548659660      1
#> 922   1.3229193512      1
#> 923   1.8378168546      1
#> 924   1.3693921474      1
#> 925   2.3127181660      1
#> 926  -1.4429010811      1
#> 927   1.4088413176      1
#> 928   1.0274693569      1
#> 929   1.3010793752      1
#> 930   0.3003404966      1
#> 931   1.6626482557      1
#> 932   1.1967564510      1
#> 933   0.8924796807      1
#> 934   0.2306768279      1
#> 935  -0.3516142440      1
#> 936   0.5282918059      1
#> 937  -0.7959919182      1
#> 938   1.4449522508      1
#> 939  -1.0734167887      1
#> 940   3.4809486062      1
#> 941  -0.6138376401      1
#> 942  -0.7703049559      1
#> 943   0.9973203829      1
#> 944  -0.0309107025      1
#> 945  -0.6715503990      1
#> 946   1.5340713990      1
#> 947   1.2389034465      1
#> 948  -0.1104557574      1
#> 949   2.7845598375      1
#> 950   0.2362503281      1
#> 951  -0.5505820556      1
#> 952   1.6725023233      1
#> 953   3.2784338184      1
#> 954   2.7232633324      1
#> 955  -0.1796767040      1
#> 956   0.0571043738      1
#> 957  -0.2000074913      1
#> 958   0.9099026411      1
#> 959   1.0120975163      1
#> 960   0.5402955088      1
#> 961  -0.4701338535      1
#> 962   1.3302770019      1
#> 963   0.9723579196      1
#> 964   0.0987467314      1
#> 965  -0.8333053253      1
#> 966   0.8862894038      1
#> 967   1.1951004117      1
#> 968   0.3063338031      1
#> 969   1.1445530797      1
#> 970   1.1311349918      1
#> 971   1.2073891112      1
#> 972  -0.2873182028      1
#> 973   0.8756238989      1
#> 974   0.6555706629      1
#> 975  -0.4348313761      1
#> 976  -0.5771665858      1
#> 977  -0.4067345310      1
#> 978  -0.5769972745      1
#> 979   1.4743327499      1
#> 980   1.0402053309      1
#> 981   3.1058645240      1
#> 982   1.6928884329      1
#> 983  -0.2935585352      1
#> 984   0.0587745816      1
#> 985   1.6071948887      1
#> 986   2.4055667440      1
#> 987   1.0613013868      1
#> 988   1.6382600744      1
#> 989   1.1043333503      1
#> 990  -0.4385987370      1
#> 991   0.8510338097      1
#> 992   0.1409932106      1
#> 993  -0.4079174378      1
#> 994   2.2907761162      1
#> 995   0.8039745141      1
#> 996   2.3646948987      1
#> 997   1.3889373503      1
#> 998   2.1960802480      1
#> 999   1.7635559437      1
#> 1000  0.5575001953      1