Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -2.3724447010      0
#> 2     1.4236135851      0
#> 3    -1.1007535827      0
#> 4     1.2830434814      0
#> 5     0.3169315015      0
#> 6    -0.2447795954      0
#> 7    -0.7424077800      0
#> 8    -0.9584922729      0
#> 9     0.3814739095      0
#> 10    0.0665991576      0
#> 11    0.1104879397      0
#> 12   -0.2875878430      0
#> 13   -0.3726207666      0
#> 14    1.2076788071      0
#> 15   -0.4895425759      0
#> 16   -0.1755264917      0
#> 17   -1.7308888373      0
#> 18    0.6037357364      0
#> 19   -0.1798565043      0
#> 20    0.4843191133      0
#> 21    0.4797939533      0
#> 22    0.5986714684      0
#> 23   -0.4990983655      0
#> 24    0.6976967338      0
#> 25    0.6507360802      0
#> 26   -0.2603368933      0
#> 27   -1.0511100936      0
#> 28   -0.9725625601      0
#> 29    0.8752830692      0
#> 30   -0.6090827056      0
#> 31    1.1775883018      0
#> 32   -0.6170529388      0
#> 33   -0.0605023829      0
#> 34   -0.4050545355      0
#> 35   -1.0218913043      0
#> 36    0.1076899750      0
#> 37   -0.1827401352      0
#> 38   -0.6231665122      0
#> 39   -2.1635254023      0
#> 40    0.7469345466      0
#> 41    0.5001129006      0
#> 42   -0.6435258669      0
#> 43    1.7612127931      0
#> 44   -0.6335634409      0
#> 45   -1.1216553081      0
#> 46   -1.6242762839      0
#> 47   -0.2216867955      0
#> 48    1.8172211692      0
#> 49   -1.7323201313      0
#> 50   -0.1071955077      0
#> 51   -0.1281640772      0
#> 52   -0.7419621709      0
#> 53   -0.4502390930      0
#> 54   -0.6932420078      0
#> 55   -0.6298366337      0
#> 56    0.3258604655      0
#> 57    1.0050876281      0
#> 58    0.0815903370      0
#> 59   -0.0621398138      0
#> 60   -1.0901683816      0
#> 61    0.4177397597      0
#> 62    1.8943832689      0
#> 63   -0.1599914497      0
#> 64   -0.0808962207      0
#> 65   -0.9510154619      0
#> 66   -0.7421124988      0
#> 67   -1.3892529558      0
#> 68    1.2182489219      0
#> 69    1.6595749031      0
#> 70   -1.0056777060      0
#> 71   -0.2296326158      0
#> 72    1.1622051371      0
#> 73    0.7542299759      0
#> 74    2.3532810841      0
#> 75    0.6318277672      0
#> 76    1.1902290960      0
#> 77   -1.6580411330      0
#> 78   -0.4868429163      0
#> 79    0.4149291359      0
#> 80   -1.4265037282      0
#> 81    0.2554625440      0
#> 82   -0.1140029190      0
#> 83   -1.4571631381      0
#> 84    0.4481768063      0
#> 85    1.3126131159      0
#> 86   -0.5925103000      0
#> 87   -0.5936799614      0
#> 88    0.6373999705      0
#> 89   -0.9575773060      0
#> 90    0.8081752139      0
#> 91   -0.2415705655      0
#> 92   -0.4252652887      0
#> 93   -1.9203164887      0
#> 94    0.0830603870      0
#> 95   -0.6665398262      0
#> 96   -0.3555175740      0
#> 97   -1.6797506054      0
#> 98   -0.0180387858      0
#> 99    0.6422265962      0
#> 100   0.2206890563      0
#> 101   0.2412127346      0
#> 102  -2.1801591910      0
#> 103   0.4682805114      0
#> 104  -1.9285010722      0
#> 105  -1.1237852025      0
#> 106   0.1509200227      0
#> 107   2.6984783320      0
#> 108  -1.9334508305      0
#> 109   0.0756950715      0
#> 110   0.1285495265      0
#> 111  -1.0371086799      0
#> 112  -0.6799894663      0
#> 113  -0.0154040560      0
#> 114  -1.1791730590      0
#> 115  -0.0758165420      0
#> 116   0.1292697563      0
#> 117  -1.5811775774      0
#> 118   0.5105175657      0
#> 119   0.7606763102      0
#> 120   0.6905923398      0
#> 121  -0.5951391082      0
#> 122   1.2848564184      0
#> 123  -0.8023286887      0
#> 124   0.1615913126      0
#> 125  -0.5107899742      0
#> 126   0.5009782088      0
#> 127   0.1582964764      0
#> 128   0.0671837584      0
#> 129   0.7774275698      0
#> 130   0.3764274070      0
#> 131   0.3049002639      0
#> 132   1.2800944424      0
#> 133  -0.1720408226      0
#> 134  -0.2842835584      0
#> 135  -0.1521096313      0
#> 136   0.5448609153      0
#> 137  -0.3203930987      0
#> 138   0.4484533548      0
#> 139  -0.3845978172      0
#> 140  -0.6283855784      0
#> 141   1.0969071673      0
#> 142  -1.7441917976      0
#> 143  -1.2656679504      0
#> 144  -1.9913718998      0
#> 145  -0.8438763744      0
#> 146   0.9583690705      0
#> 147   0.9424760881      0
#> 148   0.3097388442      0
#> 149  -0.7170331616      0
#> 150   0.4294856815      0
#> 151  -0.8207779072      0
#> 152   0.2932095097      0
#> 153   0.3917182529      0
#> 154   1.1403273413      0
#> 155   0.8266604348      0
#> 156   0.1026568552      0
#> 157   0.6169830857      0
#> 158   1.0127400251      0
#> 159  -1.8691474995      0
#> 160  -0.0435034579      0
#> 161   0.7532680892      0
#> 162   1.1597065364      0
#> 163   0.0153531766      0
#> 164  -1.1220935809      0
#> 165   0.2384593463      0
#> 166   0.6375280013      0
#> 167  -0.2260337561      0
#> 168   0.3080372094      0
#> 169   1.1670463480      0
#> 170  -0.1355557695      0
#> 171   0.4028584231      0
#> 172  -0.0517413643      0
#> 173  -0.5917084855      0
#> 174  -0.2513219465      0
#> 175   1.6758581678      0
#> 176  -1.3724635257      0
#> 177  -0.8396970069      0
#> 178   1.0237396954      0
#> 179  -0.8903687372      0
#> 180  -0.6760042722      0
#> 181   0.0418229999      0
#> 182  -0.3354029813      0
#> 183   0.8796024021      0
#> 184  -0.6861055185      0
#> 185  -0.0826628625      0
#> 186   0.0910813773      0
#> 187  -0.7463346980      0
#> 188   1.1705154106      0
#> 189  -1.4497767832      0
#> 190   0.4354731640      0
#> 191   0.6271856783      0
#> 192   0.7669776974      0
#> 193  -0.4622272301      0
#> 194  -1.5498876963      0
#> 195  -0.6996067091      0
#> 196   1.9887970209      0
#> 197  -0.9428553477      0
#> 198  -0.1257992295      0
#> 199  -0.5400897302      0
#> 200  -0.1567516652      0
#> 201   0.4167096845      0
#> 202  -0.7344759129      0
#> 203   0.4685857600      0
#> 204  -0.3073528064      0
#> 205  -2.5613575868      0
#> 206   0.4566048517      0
#> 207   0.1099762927      0
#> 208  -1.9710125417      0
#> 209  -1.2497339820      0
#> 210   0.6078053900      0
#> 211  -0.5336430596      0
#> 212   1.1289655367      0
#> 213   0.5727449851      0
#> 214   0.3403017906      0
#> 215  -0.1098758413      0
#> 216   1.0755594449      0
#> 217   1.3454819184      0
#> 218   0.6726851792      0
#> 219   0.5692000692      0
#> 220  -0.8086900000      0
#> 221  -0.6269496280      0
#> 222   1.2367868070      0
#> 223  -0.0816563845      0
#> 224  -1.1536258736      0
#> 225  -0.5762451284      0
#> 226  -0.5859162433      0
#> 227  -0.9594523452      0
#> 228   0.7385932123      0
#> 229  -0.8853759240      0
#> 230  -0.0691953795      0
#> 231  -0.8823580804      0
#> 232  -1.3886777965      0
#> 233   0.4619275452      0
#> 234   1.9475632778      0
#> 235   0.2479887621      0
#> 236  -0.0963727518      0
#> 237   0.7293279076      0
#> 238  -1.1703812163      0
#> 239   0.9522040395      0
#> 240  -0.9144299644      0
#> 241   0.2219708095      0
#> 242  -0.3684463819      0
#> 243  -0.6315033176      0
#> 244   0.8920167115      0
#> 245   2.7443646261      0
#> 246  -1.2867775936      0
#> 247  -0.0262150795      0
#> 248  -1.2840477749      0
#> 249   2.2022262668      0
#> 250  -0.8436047146      0
#> 251   1.5259087389      0
#> 252   0.7942974749      0
#> 253  -0.9415513715      0
#> 254  -0.0636323540      0
#> 255   1.3409608889      0
#> 256  -0.0391835586      0
#> 257  -2.9243995100      0
#> 258   0.0113421020      0
#> 259   1.4696056822      0
#> 260  -0.9094279073      0
#> 261   0.3181200731      0
#> 262   0.8433114465      0
#> 263  -2.0323499108      0
#> 264   1.7632154905      0
#> 265   0.5913517200      0
#> 266   1.3761009595      0
#> 267  -0.9584429026      0
#> 268  -1.9883343611      0
#> 269   0.9055870920      0
#> 270  -0.0021804650      0
#> 271   0.0315605227      0
#> 272  -1.4610995684      0
#> 273  -1.9207625096      0
#> 274   0.3981065934      0
#> 275   1.3135018855      0
#> 276   0.4893996885      0
#> 277  -0.0924411722      0
#> 278  -1.3935816067      0
#> 279   0.7817070850      0
#> 280  -0.8685928961      0
#> 281  -0.5273903482      0
#> 282  -0.3652037352      0
#> 283   1.5703645494      0
#> 284  -0.2071610634      0
#> 285  -0.9975724066      0
#> 286  -0.8743832943      0
#> 287   0.3945179214      0
#> 288  -0.5181901468      0
#> 289  -0.5603692251      0
#> 290  -2.3728270061      0
#> 291  -0.6154684086      0
#> 292   1.0669307238      0
#> 293   0.3736501428      0
#> 294   1.3207915921      0
#> 295   0.6804483600      0
#> 296   0.9888833384      0
#> 297   0.1017410629      0
#> 298  -0.9710336147      0
#> 299  -0.9377658976      0
#> 300   0.3667944193      0
#> 301  -0.3053504485      0
#> 302  -0.0004052592      0
#> 303   1.3575616278      0
#> 304   0.1749410536      0
#> 305   2.5543492600      0
#> 306   1.7941504634      0
#> 307   0.5386293626      0
#> 308   0.6228401866      0
#> 309  -2.1977767052      0
#> 310   0.9372364620      0
#> 311  -1.0055135539      0
#> 312  -1.1300180068      0
#> 313   0.9705172144      0
#> 314   0.8264709907      0
#> 315   1.9111356031      0
#> 316   1.1185725228      0
#> 317  -0.2329480313      0
#> 318  -0.2521246214      0
#> 319  -0.0580162168      0
#> 320   0.4938513758      0
#> 321  -0.8629463572      0
#> 322   1.1997135469      0
#> 323   0.1466491873      0
#> 324   1.3860733662      0
#> 325   1.2523237564      0
#> 326   1.6854028034      0
#> 327  -0.3268885281      0
#> 328   0.0255958266      0
#> 329  -0.2259344983      0
#> 330   0.8378879650      0
#> 331   0.3403968044      0
#> 332  -0.1858022669      0
#> 333  -0.4336975364      0
#> 334  -0.5122083415      0
#> 335  -0.6962395354      0
#> 336  -0.1367055888      0
#> 337   0.8876874533      0
#> 338   0.5413643878      0
#> 339  -0.8480321423      0
#> 340  -0.2688097102      0
#> 341   0.9937324731      0
#> 342   0.0950141957      0
#> 343   0.3621453096      0
#> 344   0.1778978157      0
#> 345   3.2276798493      0
#> 346   1.3333079393      0
#> 347   0.2350700436      0
#> 348   1.0289312649      0
#> 349  -0.1121107205      0
#> 350   0.3300908367      0
#> 351   0.1790428192      0
#> 352   0.6747907007      0
#> 353   0.9154515629      0
#> 354   0.0309813202      0
#> 355  -0.9345967355      0
#> 356  -0.3425272070      0
#> 357  -2.1991019238      0
#> 358   0.7667958207      0
#> 359  -0.1688678140      0
#> 360   2.0948780073      0
#> 361  -1.7486090634      0
#> 362   0.4631857127      0
#> 363   0.0251927578      0
#> 364   0.7809542560      0
#> 365  -1.5004360664      0
#> 366  -1.0973703370      0
#> 367  -1.1602555661      0
#> 368   0.3558708807      0
#> 369  -0.1728728792      0
#> 370   1.5807810447      0
#> 371   1.1640971616      0
#> 372   0.0477366697      0
#> 373   1.4438310922      0
#> 374   1.0632680104      0
#> 375   1.2112539046      0
#> 376  -0.5240553746      0
#> 377  -2.1451883227      0
#> 378  -1.1112794212      0
#> 379   1.8131861382      0
#> 380   0.1422485156      0
#> 381   1.3368797516      0
#> 382   0.3825045584      0
#> 383  -0.3822448193      0
#> 384  -0.3287288609      0
#> 385  -1.0833185056      0
#> 386   0.2909249187      0
#> 387  -0.0881889639      0
#> 388   0.7583901107      0
#> 389  -2.1423229402      0
#> 390   0.1709898738      0
#> 391  -0.7449829738      0
#> 392   0.6054861669      0
#> 393  -0.1161349669      0
#> 394  -0.5278800897      0
#> 395  -0.4838892033      0
#> 396  -0.1948608762      0
#> 397   1.4639690683      0
#> 398   0.4691375674      0
#> 399   1.3287950013      0
#> 400  -0.6010153779      0
#> 401  -0.8931470086      0
#> 402  -0.4350027119      0
#> 403  -0.1116190817      0
#> 404   0.5181011690      0
#> 405   0.1539289109      0
#> 406   1.7786504920      0
#> 407  -0.1973750883      0
#> 408   0.5095914649      0
#> 409   0.2996364165      0
#> 410  -2.2395941236      0
#> 411  -0.3607340142      0
#> 412  -1.5306341485      0
#> 413  -1.1328139158      0
#> 414   0.8126865972      0
#> 415   0.5805439856      0
#> 416   0.6956831965      0
#> 417   0.6701766413      0
#> 418  -0.5717510786      0
#> 419   0.6781375662      0
#> 420   1.4777340731      0
#> 421  -0.7677103175      0
#> 422   0.3380608217      0
#> 423   0.8097537595      0
#> 424  -0.9920470814      0
#> 425   0.1683690232      0
#> 426  -0.8401263706      0
#> 427   0.3929550999      0
#> 428  -1.1968672805      0
#> 429  -0.0352850700      0
#> 430  -0.6262158527      0
#> 431   0.4570255838      0
#> 432   0.7386584747      0
#> 433   0.9597536141      0
#> 434  -0.1185596805      0
#> 435   0.0553852704      0
#> 436  -0.5837516890      0
#> 437   0.6300805469      0
#> 438   0.2036890591      0
#> 439  -0.6745155850      0
#> 440  -0.4722853983      0
#> 441  -0.3713695156      0
#> 442   0.6320841623      0
#> 443  -0.1715619784      0
#> 444   0.8616160162      0
#> 445   0.5901127658      0
#> 446  -0.3946135616      0
#> 447  -1.4464467788      0
#> 448   1.4808774252      0
#> 449  -2.5764560862      0
#> 450   0.8161441917      0
#> 451  -0.7323862329      0
#> 452   1.3645994000      0
#> 453   0.1784499693      0
#> 454  -0.4905453982      0
#> 455  -0.1466560015      0
#> 456  -0.3352138128      0
#> 457   0.8182887053      0
#> 458  -0.6934781925      0
#> 459  -1.0502396360      0
#> 460   0.7097254158      0
#> 461   0.0425819305      0
#> 462   0.4026737450      0
#> 463   0.0523653572      0
#> 464  -0.8720612492      0
#> 465  -1.1177756091      0
#> 466  -0.2134224256      0
#> 467  -1.8528222694      0
#> 468  -0.3233349600      0
#> 469  -1.0418956773      0
#> 470  -1.3691230448      0
#> 471   0.7255244536      0
#> 472   0.8297004286      0
#> 473   1.9206384408      0
#> 474  -0.3446257942      0
#> 475  -0.4315963139      0
#> 476  -0.7144322293      0
#> 477   1.4333198169      0
#> 478  -0.3565368831      0
#> 479  -0.0134149283      0
#> 480  -0.1809232716      0
#> 481   0.0175064578      0
#> 482  -0.8828379342      0
#> 483  -0.5088498549      0
#> 484  -0.5183434111      0
#> 485   0.8025015046      0
#> 486   1.0226800896      0
#> 487   1.0665585653      0
#> 488   0.6821302899      0
#> 489  -1.7646809111      0
#> 490   1.6452076095      0
#> 491  -0.7186737721      0
#> 492  -1.4033697610      0
#> 493   1.1125150715      0
#> 494   0.1815825079      0
#> 495  -1.1892877887      0
#> 496  -0.5781648883      0
#> 497  -0.3843391127      0
#> 498   0.2825877359      0
#> 499  -0.4475498704      0
#> 500  -2.4825800928      0
#> 501   0.9242442106      0
#> 502  -0.0891151490      0
#> 503  -0.9804399211      0
#> 504  -0.1605061003      0
#> 505   0.1337409216      0
#> 506   1.1438331036      0
#> 507   0.2807107854      0
#> 508   3.0835688714      0
#> 509  -0.4794226347      0
#> 510  -1.5488039275      0
#> 511  -0.0431225270      0
#> 512  -1.5986380876      0
#> 513   0.4307745205      0
#> 514  -1.0672601851      0
#> 515  -0.3491853225      0
#> 516  -0.1298486516      0
#> 517   0.4614589096      0
#> 518   0.4241663456      0
#> 519  -1.7394860296      0
#> 520   1.4699455903      0
#> 521  -0.3811146142      0
#> 522   1.5972627050      0
#> 523  -0.7499025598      0
#> 524   0.4326773281      0
#> 525  -0.6795022903      0
#> 526   2.1968050790      0
#> 527   0.3053664576      0
#> 528   1.5859817832      0
#> 529  -0.2720978971      0
#> 530   1.1569317637      0
#> 531  -0.4835851276      0
#> 532   0.6313552622      0
#> 533  -1.2894542697      0
#> 534   0.5661658833      0
#> 535  -0.8596573841      0
#> 536  -1.2680443695      0
#> 537  -1.3106714454      0
#> 538  -0.0016758386      0
#> 539  -1.8263561720      0
#> 540  -1.2863104697      0
#> 541   0.7125802496      0
#> 542   0.6722342141      0
#> 543   0.2774214697      0
#> 544   2.2349019504      0
#> 545   1.3484703247      0
#> 546   0.0197961454      0
#> 547  -1.1722712632      0
#> 548  -0.7795087066      0
#> 549  -0.7364249235      0
#> 550   0.9958095104      0
#> 551   0.1853015839      0
#> 552   1.9741896018      0
#> 553  -0.0092257404      0
#> 554   0.3333569557      0
#> 555   0.5920612058      0
#> 556   0.3168782227      0
#> 557  -0.5668763657      0
#> 558   0.3296646404      0
#> 559  -0.4735788367      0
#> 560   0.9520723440      0
#> 561   1.1460452590      0
#> 562  -0.4034732381      0
#> 563   0.6984670375      0
#> 564  -0.3373510440      0
#> 565  -0.1790528694      0
#> 566  -1.6999879361      0
#> 567   0.5864520666      0
#> 568   0.0365599432      0
#> 569   0.3619383577      0
#> 570   0.0384309586      0
#> 571  -0.4658604286      0
#> 572   0.7772113667      0
#> 573  -1.0854341760      0
#> 574   1.0519868153      0
#> 575   1.0837962197      0
#> 576  -1.2937647262      0
#> 577  -1.8899730897      0
#> 578   1.4402442438      0
#> 579   0.9345603205      0
#> 580  -2.7707957229      0
#> 581  -1.3909364238      0
#> 582  -0.4012449023      0
#> 583   1.6880592158      0
#> 584  -0.0227856204      0
#> 585  -1.6991086656      0
#> 586   1.4992853350      0
#> 587   0.2460885951      0
#> 588  -0.6377039001      0
#> 589  -1.2236160947      0
#> 590   0.2470782524      0
#> 591   0.3959057222      0
#> 592  -0.0262157763      0
#> 593  -1.0933890178      0
#> 594   0.0058615135      0
#> 595   0.0373616660      0
#> 596  -0.3896580553      0
#> 597  -1.3679648801      0
#> 598   1.8841745168      0
#> 599   1.8827169175      0
#> 600   0.6568525831      0
#> 601  -1.1811204716      0
#> 602  -0.3393935485      0
#> 603  -0.6470919400      0
#> 604   0.7823563749      0
#> 605  -1.1278067062      0
#> 606   0.1967840759      0
#> 607  -0.2776058876      0
#> 608  -0.5631585152      0
#> 609  -0.6144081218      0
#> 610   1.4616625221      0
#> 611   0.2876858697      0
#> 612   0.0650407807      0
#> 613   1.9627698079      0
#> 614  -1.4926364575      0
#> 615   0.3795484026      0
#> 616  -0.9506972627      0
#> 617  -0.0375027812      0
#> 618   0.0218401381      0
#> 619   1.7572109006      0
#> 620  -1.3373933106      0
#> 621   1.7282045203      0
#> 622   0.2385498392      0
#> 623   0.9759705611      0
#> 624   0.7710074755      0
#> 625  -0.7037431251      0
#> 626  -1.2030994637      0
#> 627   0.9037862375      0
#> 628   1.2257304650      0
#> 629  -0.7681168828      0
#> 630  -0.7571032502      0
#> 631  -1.3120231840      0
#> 632  -2.0139129831      0
#> 633   1.8657367126      0
#> 634  -1.6619229144      0
#> 635  -0.9109052273      0
#> 636   0.1904887821      0
#> 637  -1.1948034942      0
#> 638  -1.0007187769      0
#> 639   1.0406652983      0
#> 640  -1.2357297219      0
#> 641  -1.4606094615      0
#> 642  -0.4026265022      0
#> 643   1.6684918580      0
#> 644   0.9493321017      0
#> 645  -1.1702999743      0
#> 646   0.8176828811      0
#> 647   1.2945197463      0
#> 648  -0.0917287922      0
#> 649  -1.2834734434      0
#> 650  -0.7713613938      0
#> 651   0.5404098925      0
#> 652   0.4562381039      0
#> 653  -0.6421490458      0
#> 654   2.6866717926      0
#> 655   1.0342838922      0
#> 656  -0.6726662680      0
#> 657  -0.6536887059      0
#> 658  -1.5869611132      0
#> 659   0.3210439802      0
#> 660   0.5905182354      0
#> 661  -0.6270929117      0
#> 662   0.3106957034      0
#> 663  -0.7641567578      0
#> 664   0.0911844495      0
#> 665   0.2273080509      0
#> 666  -0.6896721415      0
#> 667  -1.3660490435      0
#> 668   0.7015064059      0
#> 669  -1.4811263782      0
#> 670  -1.5947818181      0
#> 671  -0.3966715235      0
#> 672   1.0361893628      0
#> 673   1.6704263146      0
#> 674   1.0971144142      0
#> 675  -0.4123735142      0
#> 676  -0.4422026007      0
#> 677   0.6595112087      0
#> 678   0.7527611765      0
#> 679  -1.8051314758      0
#> 680  -0.0014476904      0
#> 681   1.3656333077      0
#> 682   1.0043947212      0
#> 683  -0.5973263978      0
#> 684  -0.9143180991      0
#> 685   0.3430634005      0
#> 686   1.2175772115      0
#> 687  -1.6919396328      0
#> 688   1.1236479256      0
#> 689  -0.5059381995      0
#> 690  -0.3520030108      0
#> 691  -0.7244822971      0
#> 692  -0.7027765761      0
#> 693  -0.6955476406      0
#> 694  -0.6181799037      0
#> 695  -0.5593273678      0
#> 696   0.8591535401      0
#> 697   0.7950141779      0
#> 698  -0.3013797089      0
#> 699   1.8214364490      0
#> 700  -0.5494252171      0
#> 701  -1.3646697346      0
#> 702   0.5787200487      0
#> 703   0.3530521605      0
#> 704  -0.2024628908      0
#> 705   0.4131222932      0
#> 706  -0.5938187622      0
#> 707   0.4479479570      0
#> 708   1.1833495290      0
#> 709   1.3629634935      0
#> 710   0.1602284086      0
#> 711   1.0203409467      0
#> 712   0.5600431152      0
#> 713   1.8355180461      0
#> 714   1.0211286700      0
#> 715   0.1894738970      0
#> 716  -0.4374550863      0
#> 717  -0.3164174505      0
#> 718  -0.7343518946      0
#> 719  -0.6391749141      0
#> 720   2.2170750809      0
#> 721   0.4858845989      0
#> 722  -0.5733364356      0
#> 723   0.7570394407      0
#> 724   0.2148451243      0
#> 725  -0.6299940389      0
#> 726   1.3333044296      0
#> 727  -0.5917856088      0
#> 728  -0.5824249278      0
#> 729  -0.5845662002      0
#> 730   0.9177387457      0
#> 731   1.9810733288      0
#> 732   0.0347695804      0
#> 733  -0.2677515170      0
#> 734   0.0067217735      0
#> 735  -0.7596894769      0
#> 736  -1.8387117000      0
#> 737   2.1550362909      0
#> 738   0.9537009420      0
#> 739   0.2000372028      0
#> 740  -1.2349951203      0
#> 741   0.8894002934      0
#> 742   0.1671710314      0
#> 743   0.4247802975      0
#> 744   0.2130731682      0
#> 745   0.7846589201      0
#> 746   0.1057651178      0
#> 747   0.0694733821      0
#> 748  -0.7342449093      0
#> 749   0.7576702462      0
#> 750  -2.1708237501      0
#> 751   0.7796189730      0
#> 752  -0.8298451007      0
#> 753   0.0200647171      0
#> 754  -1.4948754208      0
#> 755  -0.8165989323      0
#> 756   0.4865853271      0
#> 757  -1.9871496313      0
#> 758   0.4461707588      0
#> 759   0.7055436669      0
#> 760   1.3325137205      0
#> 761  -0.7625910201      0
#> 762   0.0101407348      0
#> 763   0.5873536675      0
#> 764  -1.7346477093      0
#> 765  -1.4230494852      0
#> 766  -1.3420156997      0
#> 767  -1.4463895350      0
#> 768   1.1188849799      0
#> 769   0.1602234591      0
#> 770   1.2226869900      0
#> 771  -2.0193144981      0
#> 772  -1.3781683629      0
#> 773   2.2540569783      0
#> 774  -0.0190570472      0
#> 775  -0.7504848805      0
#> 776   1.8196244863      0
#> 777  -1.4936801415      0
#> 778   0.0697858119      0
#> 779  -1.2429215111      0
#> 780   0.5945968444      0
#> 781  -0.2209245754      0
#> 782   1.4829588297      0
#> 783  -0.7118339939      0
#> 784  -0.1742083773      0
#> 785  -0.8147189754      0
#> 786  -1.6979373696      0
#> 787   0.1546284789      0
#> 788   0.5441313113      0
#> 789  -0.1153542047      0
#> 790   0.5341306050      0
#> 791   1.2874791100      0
#> 792  -0.5904188712      0
#> 793   1.0213581772      0
#> 794  -0.4931566949      0
#> 795   0.1335028971      0
#> 796  -0.1546488549      0
#> 797  -0.0447910677      0
#> 798  -1.1774699260      0
#> 799   0.1733823575      0
#> 800   0.1582081766      0
#> 801  -0.7882572585      0
#> 802  -0.8003439884      0
#> 803  -0.8809433997      0
#> 804  -0.2534456089      0
#> 805  -0.7010461057      0
#> 806  -1.2392760244      0
#> 807   1.1259200349      0
#> 808  -0.9836085979      0
#> 809  -1.2410788891      0
#> 810   0.2702017293      0
#> 811  -0.2011105246      0
#> 812   0.3244689945      0
#> 813  -0.3830907477      0
#> 814   1.6201323762      0
#> 815   0.9006237996      0
#> 816  -0.1606566953      0
#> 817   1.0528180699      0
#> 818  -0.0944065742      0
#> 819  -1.2114819809      0
#> 820   1.1470927607      0
#> 821  -1.3039453715      0
#> 822   1.3002043233      0
#> 823   0.2563980486      0
#> 824   0.2190323196      0
#> 825   0.5463735097      0
#> 826   0.6832173898      0
#> 827   0.2721745326      0
#> 828   0.1363028596      0
#> 829  -0.0521637856      0
#> 830   0.0606238680      0
#> 831   0.8406347847      0
#> 832  -0.5398997560      0
#> 833   0.4449193907      0
#> 834   1.0679724690      0
#> 835   0.4984026097      0
#> 836  -0.7489353599      0
#> 837  -0.2295026430      0
#> 838   0.7304603339      0
#> 839   0.7717401313      0
#> 840  -1.6193853403      0
#> 841   0.9839099431      0
#> 842  -0.1518351733      0
#> 843  -2.1475875443      0
#> 844   1.2090598158      0
#> 845   0.5491312896      0
#> 846  -0.4332876724      0
#> 847   1.8374811394      0
#> 848  -0.7792405997      0
#> 849  -1.3634635927      0
#> 850  -0.4824767757      0
#> 851   0.8787813059      0
#> 852  -0.1813567168      0
#> 853   0.4732171925      0
#> 854   0.2295468852      0
#> 855   0.7422309000      0
#> 856  -0.1245706094      0
#> 857  -0.5359312464      0
#> 858   2.5225457750      0
#> 859   0.1956315021      0
#> 860   0.4936742243      0
#> 861   1.5838169568      0
#> 862  -1.5265339321      0
#> 863  -0.6554547440      0
#> 864   1.0686875789      0
#> 865  -2.1131808063      0
#> 866   0.6464703031      0
#> 867  -0.7572387350      0
#> 868   1.2255905218      0
#> 869  -0.1290497977      0
#> 870  -1.5494458923      0
#> 871   0.5462504733      0
#> 872   1.0828191843      0
#> 873  -0.8444427253      0
#> 874  -0.0740983667      0
#> 875   0.1260522842      0
#> 876   0.6978720171      0
#> 877  -0.6584847064      0
#> 878   0.1650928373      0
#> 879  -0.8147581751      0
#> 880  -0.3734557706      0
#> 881  -0.5963495125      0
#> 882  -2.6322970341      0
#> 883   0.0336907093      0
#> 884   1.1142799497      0
#> 885   2.6316776788      0
#> 886   0.5776505265      0
#> 887  -1.7715230838      0
#> 888   0.4604052235      0
#> 889   0.2871963795      0
#> 890  -0.9760159832      0
#> 891   0.2655493890      0
#> 892  -1.6424713399      0
#> 893  -1.6116713023      0
#> 894  -2.4534889005      0
#> 895   0.3157525876      0
#> 896   0.9218814923      0
#> 897   2.2435624571      0
#> 898  -0.1910500091      0
#> 899  -2.6262386672      0
#> 900   1.3814637660      0
#> 901  -0.7141510154      0
#> 902  -0.4721015201      0
#> 903  -0.7674086759      0
#> 904  -0.0532890415      0
#> 905   0.6863980575      0
#> 906  -0.0071103046      0
#> 907   1.6839397401      0
#> 908   1.1741078544      0
#> 909   0.4692766587      1
#> 910   0.0361446283      1
#> 911  -0.3279254781      1
#> 912   0.9416307080      1
#> 913  -0.0156914034      1
#> 914   0.6409335112      1
#> 915  -0.2278043795      1
#> 916   0.1354797529      1
#> 917   0.4672656485      1
#> 918   1.2783807404      1
#> 919   0.3114295277      1
#> 920  -2.5672279877      1
#> 921  -1.2667989402      1
#> 922   2.4414961183      1
#> 923   0.3258934860      1
#> 924   0.7476946790      1
#> 925   1.8634675650      1
#> 926   0.7079365781      1
#> 927   1.7919500826      1
#> 928   0.8777285615      1
#> 929   2.0705718161      1
#> 930   1.4644577775      1
#> 931   1.2364282612      1
#> 932   1.2765972648      1
#> 933  -0.6572331182      1
#> 934   0.9014458054      1
#> 935   1.8561904635      1
#> 936   1.7433849234      1
#> 937   1.4817968984      1
#> 938  -0.3316089063      1
#> 939   1.6365816923      1
#> 940  -0.8485091104      1
#> 941   1.8421272713      1
#> 942   0.2127077939      1
#> 943  -0.1857172159      1
#> 944  -0.3630618190      1
#> 945  -0.0139203745      1
#> 946   1.2540772929      1
#> 947   2.7048285049      1
#> 948   0.4899383898      1
#> 949   1.4884362075      1
#> 950   0.5083618752      1
#> 951  -1.0654986774      1
#> 952   0.5789225581      1
#> 953  -2.1514272149      1
#> 954   2.5078422782      1
#> 955   1.9739089134      1
#> 956   1.6025108790      1
#> 957   2.2680762140      1
#> 958   1.9281803311      1
#> 959   1.4260502402      1
#> 960   2.6410001994      1
#> 961   0.1233600561      1
#> 962  -0.6214396057      1
#> 963   0.5543183127      1
#> 964   0.5765820083      1
#> 965   0.6751566608      1
#> 966   0.8638347230      1
#> 967   1.4008997727      1
#> 968   1.4635835750      1
#> 969   1.0377757057      1
#> 970   0.4168790917      1
#> 971   0.7775853566      1
#> 972   0.4969681644      1
#> 973   1.0779921179      1
#> 974   0.4274345907      1
#> 975   1.2496246021      1
#> 976  -1.0487333453      1
#> 977   0.9152203176      1
#> 978   1.2528371434      1
#> 979   0.3439944471      1
#> 980   2.0341506498      1
#> 981  -0.6528944435      1
#> 982  -0.8239901549      1
#> 983  -0.4941876739      1
#> 984   1.3190570625      1
#> 985  -0.0315600707      1
#> 986   3.3660088299      1
#> 987   1.6693669057      1
#> 988   1.3068713306      1
#> 989   0.8680556100      1
#> 990  -0.1169823867      1
#> 991  -1.4970622919      1
#> 992   0.7931631383      1
#> 993  -0.1722555903      1
#> 994   0.5186090003      1
#> 995  -0.2441511830      1
#> 996  -0.7994746092      1
#> 997   0.6743804251      1
#> 998   2.0595881481      1
#> 999   2.0142810821      1
#> 1000  1.3292366086      1