Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -0.1833244901      0
#> 2     0.0567430637      0
#> 3    -0.0011482726      0
#> 4    -0.8680143559      0
#> 5     0.4011578028      0
#> 6     0.5879485018      0
#> 7     0.8673411115      0
#> 8    -0.0304170957      0
#> 9    -0.1602153888      0
#> 10    0.3341938441      0
#> 11    1.7553933426      0
#> 12    0.2427545396      0
#> 13   -1.4898613920      0
#> 14    1.5122174435      0
#> 15    1.3019860234      0
#> 16    1.5511773845      0
#> 17   -0.3587727491      0
#> 18    0.0739538904      0
#> 19   -0.8685868082      0
#> 20   -0.7026323576      0
#> 21   -0.7697299751      0
#> 22    0.7833993764      0
#> 23    1.8577990008      0
#> 24    1.4799189924      0
#> 25   -0.2481210485      0
#> 26    0.3323779614      0
#> 27   -0.2816339130      0
#> 28    0.1172041817      0
#> 29   -0.8225998965      0
#> 30    0.9011571617      0
#> 31   -0.1849049314      0
#> 32    1.2326026721      0
#> 33   -0.5874580402      0
#> 34    2.7131612476      0
#> 35    0.8927546387      0
#> 36   -1.8911613470      0
#> 37    0.0949989731      0
#> 38    1.7225568816      0
#> 39   -0.6697695739      0
#> 40   -0.3227999117      0
#> 41   -0.8118851940      0
#> 42   -0.1741916563      0
#> 43    0.2099770350      0
#> 44    0.1486098008      0
#> 45    0.0600686787      0
#> 46    0.6168883316      0
#> 47    0.6612434660      0
#> 48    1.2034062499      0
#> 49    0.4052563519      0
#> 50   -1.3874520477      0
#> 51   -0.3781332366      0
#> 52    0.4237394296      0
#> 53   -2.0326401229      0
#> 54    0.0404571600      0
#> 55   -1.9541057913      0
#> 56   -0.0385320420      0
#> 57    0.9797143689      0
#> 58    0.1034744210      0
#> 59    1.4835106945      0
#> 60   -0.6099870791      0
#> 61    1.6741263959      0
#> 62    0.4833594976      0
#> 63   -1.2330791153      0
#> 64    1.2188887270      0
#> 65    1.2336626405      0
#> 66   -0.2436702040      0
#> 67   -1.4620443240      0
#> 68    0.2021514046      0
#> 69    0.1259177004      0
#> 70   -0.8825931303      0
#> 71    0.8296962683      0
#> 72   -2.0050032314      0
#> 73    0.1384109945      0
#> 74   -1.1596088700      0
#> 75   -1.6889569761      0
#> 76    2.0027955266      0
#> 77    0.5652233632      0
#> 78    1.1118032179      0
#> 79   -0.0556180512      0
#> 80   -0.6410617624      0
#> 81    1.1176358096      0
#> 82   -0.1118628288      0
#> 83   -0.0112721844      0
#> 84   -0.1979746368      0
#> 85    0.1601434077      0
#> 86    1.1445542810      0
#> 87    1.6204803764      0
#> 88   -0.4203252514      0
#> 89   -2.2765802848      0
#> 90    1.1007828762      0
#> 91    0.9178255583      0
#> 92   -0.9717005708      0
#> 93    0.0891171777      0
#> 94   -0.7293670380      0
#> 95    0.8354273166      0
#> 96   -1.5461546320      0
#> 97    2.1306123248      0
#> 98    0.3342352533      0
#> 99   -1.3583830373      0
#> 100   0.2711404801      0
#> 101  -0.9861516707      0
#> 102   0.6575630861      0
#> 103   0.6115822530      0
#> 104  -1.2089318620      0
#> 105   1.8529090595      0
#> 106  -1.2915156520      0
#> 107  -0.8583363692      0
#> 108   0.9039859196      0
#> 109   0.1854200742      0
#> 110   1.4801823049      0
#> 111   0.8216532450      0
#> 112  -0.9853150595      0
#> 113  -1.7780516683      0
#> 114  -1.0133265245      0
#> 115  -1.1017911637      0
#> 116   0.6083148645      0
#> 117  -1.0630385329      0
#> 118  -0.4695711290      0
#> 119   0.6039721580      0
#> 120   0.3109721509      0
#> 121   0.7349595838      0
#> 122  -0.0324951570      0
#> 123   0.0385032724      0
#> 124   0.7515095265      0
#> 125   1.8974248479      0
#> 126  -0.3240576527      0
#> 127   0.3360484827      0
#> 128  -0.6076857013      0
#> 129   0.1823939234      0
#> 130   1.1621471324      0
#> 131   0.5592697947      0
#> 132  -0.3230516568      0
#> 133  -0.8974038937      0
#> 134   0.8038266105      0
#> 135   0.9191512647      0
#> 136  -0.0448947901      0
#> 137  -0.8155678514      0
#> 138  -1.1062992159      0
#> 139   0.1683642903      0
#> 140  -1.6081832020      0
#> 141  -1.1020245499      0
#> 142   2.0584145336      0
#> 143  -0.0357078339      0
#> 144  -2.3405343545      0
#> 145   0.7644599753      0
#> 146   1.7815714937      0
#> 147  -0.1618048616      0
#> 148   0.7939884787      0
#> 149   0.1117955152      0
#> 150   0.8234552039      0
#> 151   0.2521888844      0
#> 152   0.6876484797      0
#> 153   0.1463467343      0
#> 154  -0.1333615918      0
#> 155   0.8868152884      0
#> 156   0.7288513765      0
#> 157   1.3619336371      0
#> 158  -0.2510761061      0
#> 159  -0.4406365270      0
#> 160  -0.9846141200      0
#> 161  -0.2566468877      0
#> 162  -1.2348650712      0
#> 163  -2.0928447169      0
#> 164  -0.4793376009      0
#> 165  -0.0839518331      0
#> 166  -1.5366220149      0
#> 167  -0.2978470975      0
#> 168   1.4602308272      0
#> 169  -0.3665166740      0
#> 170  -0.2998819358      0
#> 171  -0.9792747896      0
#> 172   1.7041356142      0
#> 173  -0.3181953204      0
#> 174  -0.9114393772      0
#> 175  -0.4685513710      0
#> 176   0.5236066089      0
#> 177  -1.7564972898      0
#> 178  -1.3689827535      0
#> 179  -1.0807257307      0
#> 180   0.7031397837      0
#> 181   0.9614708497      0
#> 182  -0.7446472991      0
#> 183  -0.6302870439      0
#> 184   1.1784144494      0
#> 185  -1.3066414293      0
#> 186  -1.8640961419      0
#> 187   0.0932598020      0
#> 188   0.8273322781      0
#> 189  -0.3867183936      0
#> 190  -0.6090268255      0
#> 191  -0.0849348423      0
#> 192   1.2612538700      0
#> 193   0.6041393757      0
#> 194  -0.3796463240      0
#> 195   1.5415681081      0
#> 196  -0.1339731771      0
#> 197   2.8613337265      0
#> 198  -0.0003438514      0
#> 199   0.8449264954      0
#> 200   0.6114848555      0
#> 201   0.2232914493      0
#> 202   0.7030199486      0
#> 203   0.4808850534      0
#> 204   2.5043015972      0
#> 205   0.2013004455      0
#> 206   0.6443710423      0
#> 207   0.8691526890      0
#> 208   0.1987967049      0
#> 209  -0.6830217052      0
#> 210   0.3538696094      0
#> 211   0.4927321605      0
#> 212  -0.0714903013      0
#> 213  -0.2860218915      0
#> 214  -0.0536179778      0
#> 215   0.3579705005      0
#> 216  -1.9393157828      0
#> 217   0.6057160156      0
#> 218  -0.0646064018      0
#> 219   1.5478977151      0
#> 220  -0.7677790973      0
#> 221  -0.9705492829      0
#> 222   0.3478753056      0
#> 223  -0.3025086052      0
#> 224  -0.3915002640      0
#> 225   0.6198612950      0
#> 226  -1.7060818084      0
#> 227  -0.3546207825      0
#> 228   1.4651393245      0
#> 229   0.6913055982      0
#> 230   1.3941081119      0
#> 231   0.4017094516      0
#> 232   2.2338419967      0
#> 233  -1.1233273373      0
#> 234  -1.1705565284      0
#> 235  -0.2251029750      0
#> 236  -0.2871453163      0
#> 237   1.3352515256      0
#> 238  -0.3515269563      0
#> 239   0.9679305607      0
#> 240   1.4254402501      0
#> 241   0.4144905024      0
#> 242  -0.5852325364      0
#> 243   1.2582997034      0
#> 244   0.4922071264      0
#> 245  -0.9919813169      0
#> 246   1.0980343455      0
#> 247   0.0269147920      0
#> 248  -0.8345358604      0
#> 249   0.2148749708      0
#> 250  -2.5497626268      0
#> 251   0.1400347424      0
#> 252   1.0206877341      0
#> 253   0.6948927718      0
#> 254   0.3029746916      0
#> 255   2.6384257391      0
#> 256   0.0588966049      0
#> 257   0.0038909530      0
#> 258  -1.7601944128      0
#> 259   0.4050767494      0
#> 260   0.1353178874      0
#> 261   1.9204605992      0
#> 262   0.4135697809      0
#> 263   0.4755189603      0
#> 264   0.9682532203      0
#> 265  -0.2532836021      0
#> 266   1.3561708063      0
#> 267  -0.0438559361      0
#> 268   0.5554267707      0
#> 269  -0.2892790636      0
#> 270   0.8163833716      0
#> 271   1.2447545904      0
#> 272   0.2831561962      0
#> 273   0.8362063206      0
#> 274  -0.5224435465      0
#> 275   0.8420070738      0
#> 276   0.5487692002      0
#> 277   0.4371924893      0
#> 278  -0.3133936444      0
#> 279  -0.1621248358      0
#> 280  -0.7532100332      0
#> 281  -0.7471685777      0
#> 282   0.1141499864      0
#> 283   2.8061289611      0
#> 284  -0.3505058170      0
#> 285  -1.1533708920      0
#> 286   2.1490690812      0
#> 287  -0.2124729617      0
#> 288  -0.2939072266      0
#> 289  -0.3307735761      0
#> 290  -0.3279906392      0
#> 291  -0.3713146949      0
#> 292   0.3860447259      0
#> 293   0.0142008806      0
#> 294  -0.1381737641      0
#> 295  -0.4919534404      0
#> 296  -1.1697595460      0
#> 297   1.6886918014      0
#> 298   0.4228066255      0
#> 299   0.0994641349      0
#> 300  -0.6318740258      0
#> 301   0.9766862517      0
#> 302   1.6549238746      0
#> 303   0.6968967912      0
#> 304   0.2164925278      0
#> 305  -0.4539991322      0
#> 306  -0.2297453895      0
#> 307   0.5540887476      0
#> 308   1.1297958379      0
#> 309   1.4622222858      0
#> 310  -0.7360947970      0
#> 311  -1.4542088479      0
#> 312  -1.1650234025      0
#> 313  -0.3753786256      0
#> 314   0.6084666842      0
#> 315   1.0660449822      0
#> 316   0.9599442401      0
#> 317  -0.3383845528      0
#> 318  -0.6718683419      0
#> 319  -1.0743991812      0
#> 320  -0.6803109672      0
#> 321  -0.8172645300      0
#> 322  -0.6742316695      0
#> 323   1.0127477009      0
#> 324  -1.2843437019      0
#> 325   0.2249499719      0
#> 326  -1.8169422428      0
#> 327  -1.0608676765      0
#> 328   0.2710552172      0
#> 329  -0.5889071426      0
#> 330  -0.5549928329      0
#> 331  -0.2241514698      0
#> 332  -1.5284224983      0
#> 333   0.3839682768      0
#> 334   0.6530217044      0
#> 335  -1.6442160113      0
#> 336  -0.3027401866      0
#> 337  -0.2529074635      0
#> 338   1.0869314584      0
#> 339   0.4752867536      0
#> 340   0.0127398832      0
#> 341   0.0528360933      0
#> 342   0.4244162081      0
#> 343  -1.5191170983      0
#> 344   0.6164324982      0
#> 345   1.1426790835      0
#> 346   1.6641298981      0
#> 347  -1.2508553219      0
#> 348  -0.5820808765      0
#> 349   0.0482044789      0
#> 350   2.8148989750      0
#> 351   0.8427826864      0
#> 352   1.2107692862      0
#> 353   2.6720181725      0
#> 354   1.6279098813      0
#> 355  -0.5021467457      0
#> 356  -0.4482083631      0
#> 357  -0.2739157541      0
#> 358  -1.5475289987      0
#> 359   0.7244880077      0
#> 360   1.1991600081      0
#> 361  -0.3413930959      0
#> 362  -0.0197146137      0
#> 363  -1.6446583143      0
#> 364   0.3149585134      0
#> 365   2.0875209419      0
#> 366   1.4019631470      0
#> 367   1.3371252194      0
#> 368   0.3841613358      0
#> 369  -0.0452991795      0
#> 370  -0.0822097446      0
#> 371   0.3206109484      0
#> 372  -1.2454059738      0
#> 373  -0.6594605912      0
#> 374   0.8971757559      0
#> 375  -1.3517170157      0
#> 376   0.7711371856      0
#> 377   1.5741932714      0
#> 378   1.2820866059      0
#> 379  -1.0018954574      0
#> 380   0.1934172283      0
#> 381   0.6678467278      0
#> 382   0.0552217827      0
#> 383   0.8796221701      0
#> 384   2.5007087683      0
#> 385  -1.0268521492      0
#> 386   0.0753256931      0
#> 387  -0.0523006013      0
#> 388  -2.8825500205      0
#> 389   0.0920575163      0
#> 390  -1.0648381906      0
#> 391  -0.0487648329      0
#> 392  -0.6779012389      0
#> 393   2.2116910935      0
#> 394   0.6070153594      0
#> 395   0.4547033008      0
#> 396   1.1088675724      0
#> 397   0.7539289907      0
#> 398  -1.3731324524      0
#> 399  -1.2291072163      0
#> 400   0.8014230030      0
#> 401  -0.4667864311      0
#> 402  -1.6369237662      0
#> 403  -0.0846325218      0
#> 404  -0.3342305571      0
#> 405  -1.6424769399      0
#> 406  -0.4444379246      0
#> 407  -0.3335249899      0
#> 408  -1.4393274733      0
#> 409  -0.2080634021      0
#> 410   1.7819753823      0
#> 411  -0.1119308499      0
#> 412   0.7141186736      0
#> 413   0.6652397418      0
#> 414  -0.2857673593      0
#> 415  -1.9577581785      0
#> 416   0.8033429323      0
#> 417  -0.0694054944      0
#> 418   0.1995192185      0
#> 419  -0.1664083450      0
#> 420   1.2240493181      0
#> 421  -0.5049875318      0
#> 422   0.4672594901      0
#> 423  -1.6641324715      0
#> 424  -0.8351848611      0
#> 425  -0.4120962743      0
#> 426   0.8723617797      0
#> 427   0.8734534415      0
#> 428  -0.1864533969      0
#> 429   0.1413606921      0
#> 430  -1.0514092999      0
#> 431   0.6323342225      0
#> 432   0.5124916076      0
#> 433  -1.9491483826      0
#> 434   0.4192464570      0
#> 435  -2.3195102483      0
#> 436   0.9052295142      0
#> 437  -0.1067073397      0
#> 438  -1.2859946660      0
#> 439  -1.4073440283      0
#> 440   0.6174860584      0
#> 441  -1.8027831634      0
#> 442   0.0735057584      0
#> 443  -0.4081613254      0
#> 444   0.0818768137      0
#> 445  -0.1320549242      0
#> 446   0.0648459761      0
#> 447  -0.1747389567      0
#> 448  -0.1234361681      0
#> 449   1.5300844508      0
#> 450   0.3501722204      0
#> 451  -1.5154374324      0
#> 452   0.9194479888      0
#> 453  -1.0208269795      0
#> 454   1.5493461232      0
#> 455  -0.4915977792      0
#> 456  -0.1708375102      0
#> 457  -1.0796813900      0
#> 458  -0.2199800254      0
#> 459  -0.0733289522      0
#> 460  -0.3855645314      0
#> 461   1.0801076230      0
#> 462  -0.0348114187      0
#> 463   1.4734702858      0
#> 464  -0.1702212075      0
#> 465  -0.4628832955      0
#> 466   0.8254476869      0
#> 467   0.9614280987      0
#> 468   0.0755658553      0
#> 469   1.2007035801      0
#> 470  -0.4025924974      0
#> 471   0.5243810969      0
#> 472   0.2991153628      0
#> 473  -0.2334075320      0
#> 474   0.0816057684      0
#> 475  -1.6120905501      0
#> 476   1.3698644982      0
#> 477  -0.7136987690      0
#> 478  -0.4795208554      0
#> 479   0.8045147595      0
#> 480   1.1808919386      0
#> 481  -1.3131756258      0
#> 482  -1.1948275458      0
#> 483  -0.9052374696      0
#> 484  -0.8625035878      0
#> 485  -1.3494319104      0
#> 486  -0.9524961235      0
#> 487  -0.9638053838      0
#> 488  -0.4057526879      0
#> 489   0.9251034586      0
#> 490  -0.9730088690      0
#> 491  -0.3544935475      0
#> 492   0.5796228278      0
#> 493  -1.3603865614      0
#> 494   2.3525662265      0
#> 495  -0.1140485568      0
#> 496  -0.0928877854      0
#> 497  -0.3406112864      0
#> 498   1.4005652064      0
#> 499  -0.9946942238      0
#> 500   0.5110716859      0
#> 501  -0.1829365077      0
#> 502  -0.4721084896      0
#> 503   2.5263036589      0
#> 504  -1.6598948094      0
#> 505   0.1317044332      0
#> 506  -0.0695535707      0
#> 507  -0.7686808948      0
#> 508   0.4101395661      0
#> 509  -0.5788689133      0
#> 510   0.7762210205      0
#> 511   0.6281953626      0
#> 512  -0.5262420031      0
#> 513   0.5668458737      0
#> 514   1.8154204611      0
#> 515  -0.1925448697      0
#> 516   0.7045952668      0
#> 517  -0.1960085847      0
#> 518  -1.1561035125      0
#> 519  -0.1517590792      0
#> 520  -0.2065843239      0
#> 521  -1.8732242355      0
#> 522   0.7312942675      0
#> 523  -0.4729800473      0
#> 524  -0.5124252707      0
#> 525  -1.2107992115      0
#> 526  -0.2104925555      0
#> 527  -0.4038671802      0
#> 528   1.5897469894      0
#> 529   1.2374528774      0
#> 530   0.3564050959      0
#> 531  -0.2690218149      0
#> 532   0.0923752337      0
#> 533   0.3526350100      0
#> 534   0.9690996095      0
#> 535   0.6964783387      0
#> 536   1.2054679291      0
#> 537  -1.3678286621      0
#> 538   0.5348808008      0
#> 539  -1.3597834259      0
#> 540  -0.7676271301      0
#> 541   0.0022573435      0
#> 542  -0.1964122932      0
#> 543  -0.2291028571      0
#> 544   1.0538367270      0
#> 545  -1.0350155011      0
#> 546   0.4022192383      0
#> 547  -0.3272874222      0
#> 548   0.0430535068      0
#> 549  -0.6213499340      0
#> 550   0.7470231197      0
#> 551   0.4455910178      0
#> 552  -2.0768960168      0
#> 553   0.3405560029      0
#> 554  -1.6491998026      0
#> 555   0.2070775595      0
#> 556   2.7822799317      0
#> 557  -0.2352663934      0
#> 558   0.0546059632      0
#> 559  -0.2409301029      0
#> 560  -0.3288284966      0
#> 561  -0.2738397306      0
#> 562  -1.4695455735      0
#> 563  -0.1556209750      0
#> 564  -1.1286242849      0
#> 565   0.3678132329      0
#> 566  -1.2998046326      0
#> 567   0.2826940051      0
#> 568   2.0458392853      0
#> 569   0.3494290497      0
#> 570   1.1222215092      0
#> 571   1.8870521874      0
#> 572  -2.9253483884      0
#> 573   0.5342396812      0
#> 574   0.2828787422      0
#> 575   0.6333296988      0
#> 576   0.3592065869      0
#> 577  -0.0515500025      0
#> 578   0.2302576707      0
#> 579  -0.8867235850      0
#> 580  -0.1484259599      0
#> 581   0.9487542387      0
#> 582   0.3659530563      0
#> 583  -0.5033474526      0
#> 584  -0.7878064774      0
#> 585   0.9262461310      0
#> 586   0.1520942371      0
#> 587   0.4001109120      0
#> 588  -1.2277845009      0
#> 589   0.0293004857      0
#> 590   0.7622602434      0
#> 591  -2.1076476022      0
#> 592  -0.1058352645      0
#> 593  -0.4636074337      0
#> 594   0.7837623856      0
#> 595   1.2723345523      0
#> 596   0.0177140978      0
#> 597   0.1457234942      0
#> 598  -1.3309727539      0
#> 599  -0.2758526176      0
#> 600   1.3858871414      0
#> 601   0.6371179676      0
#> 602  -1.0627588317      0
#> 603   0.7977510530      0
#> 604   0.8278373288      0
#> 605   0.0667618235      0
#> 606  -0.5582538563      0
#> 607   0.9260853146      0
#> 608  -0.9105231391      0
#> 609   0.1310137172      0
#> 610   1.4571246595      0
#> 611   0.2189698285      0
#> 612   0.5185123288      0
#> 613  -0.9678174609      0
#> 614  -0.5199009455      0
#> 615  -0.6748333704      0
#> 616  -0.6198540410      0
#> 617   1.5841471745      0
#> 618   0.2337964473      0
#> 619  -0.2395647187      0
#> 620  -3.1033539240      0
#> 621   1.2158159947      0
#> 622  -0.1721173475      0
#> 623   0.6088186397      0
#> 624   1.4573283036      0
#> 625  -0.5431260894      0
#> 626   0.4191930138      0
#> 627  -0.4992800921      0
#> 628   1.5990079899      0
#> 629  -0.5676852604      0
#> 630   0.1514137971      0
#> 631  -0.7837646616      0
#> 632   0.6773997381      0
#> 633  -0.6446215559      0
#> 634   0.6645637702      0
#> 635  -1.2394833841      0
#> 636   1.1550635978      0
#> 637   0.4911118468      0
#> 638  -0.6255176951      0
#> 639   0.3400953497      0
#> 640   0.4151726718      0
#> 641   1.3422851582      0
#> 642   0.3298118478      0
#> 643   0.2721317968      0
#> 644   1.4990323772      0
#> 645  -0.1865093381      0
#> 646  -0.9420083102      0
#> 647   0.2949387838      0
#> 648   1.5051773008      0
#> 649  -1.6916953233      0
#> 650   0.2669383297      0
#> 651   1.2520714154      0
#> 652   0.1997358588      0
#> 653   0.0751640100      0
#> 654  -0.5224592627      0
#> 655  -0.1046198749      0
#> 656   3.2978353514      0
#> 657  -1.6046393123      0
#> 658  -1.6054042052      0
#> 659  -1.2982744691      0
#> 660  -1.6499709512      0
#> 661   0.6085972322      0
#> 662  -1.5234796758      0
#> 663   0.8998812229      0
#> 664  -0.8507762871      0
#> 665  -0.3538603248      0
#> 666   0.6345353912      0
#> 667   0.0937165172      0
#> 668  -0.3741345874      0
#> 669  -0.4823543149      0
#> 670   0.5666697909      0
#> 671  -1.3189715201      0
#> 672  -0.8799156445      0
#> 673   0.7162880055      0
#> 674  -0.7570068364      0
#> 675  -1.5860204872      0
#> 676  -1.0254306045      0
#> 677  -1.5719013416      0
#> 678   1.4808984505      0
#> 679   0.6171294575      0
#> 680  -0.9088656100      0
#> 681  -0.1932871235      0
#> 682   1.2505651071      0
#> 683  -0.3112114864      0
#> 684  -0.5473367915      0
#> 685   0.0934918375      0
#> 686  -1.0161888702      0
#> 687   1.4462548281      0
#> 688  -0.1090657023      0
#> 689   0.6583134629      0
#> 690   0.8534062791      0
#> 691  -0.7155307149      0
#> 692  -0.1392401678      0
#> 693  -0.5478858977      0
#> 694   0.4497931066      0
#> 695  -1.7158524606      0
#> 696  -0.2366788875      0
#> 697  -0.6342694467      0
#> 698   1.8726598312      0
#> 699   0.9164630457      0
#> 700   0.2967330382      0
#> 701  -0.4421808139      0
#> 702   0.6161062249      0
#> 703  -0.6320866808      0
#> 704  -0.7684529161      0
#> 705   0.7590378242      0
#> 706  -0.0376998316      0
#> 707  -0.9097793978      0
#> 708   0.9099416325      0
#> 709  -1.2569294344      0
#> 710  -0.7701812965      0
#> 711  -0.2169171876      0
#> 712  -0.9191116053      0
#> 713   0.4592561796      0
#> 714   0.1957924606      0
#> 715   0.5879885108      0
#> 716  -0.8575628970      0
#> 717   1.4063307471      0
#> 718  -0.0105144192      0
#> 719   0.3246604210      0
#> 720   1.3446233748      0
#> 721  -0.9695958752      0
#> 722  -1.3345511720      0
#> 723  -0.4486713042      0
#> 724  -1.0370119689      0
#> 725  -0.4898035700      0
#> 726   0.3622080868      0
#> 727  -0.4351404172      0
#> 728  -1.0556360971      0
#> 729  -0.8777522934      0
#> 730  -0.2174565802      0
#> 731  -1.2292784987      0
#> 732   0.6281548563      0
#> 733  -2.6061393599      0
#> 734  -0.5944826733      0
#> 735  -0.5569141847      0
#> 736  -1.4820512712      0
#> 737   1.9039392323      0
#> 738   1.1553737469      0
#> 739   1.9602204481      0
#> 740  -0.0438066957      0
#> 741  -0.1205927971      0
#> 742  -1.3920156794      0
#> 743   2.5705544990      0
#> 744   0.0073770170      0
#> 745   0.5790249872      0
#> 746   0.3066918231      0
#> 747   0.0996328942      0
#> 748   2.0302110777      0
#> 749  -0.2493365026      0
#> 750   0.1023221529      0
#> 751  -0.8215276634      0
#> 752   0.1054660648      0
#> 753  -0.2121618242      0
#> 754   0.1959760631      0
#> 755  -0.8703663045      0
#> 756   0.3103159280      0
#> 757   2.7379156045      0
#> 758  -1.4497946361      0
#> 759  -0.7246542740      0
#> 760  -1.3929827751      0
#> 761   0.1871468115      0
#> 762  -1.5180065327      0
#> 763  -0.2615852097      0
#> 764   1.3331514882      0
#> 765  -0.0583265501      0
#> 766   0.0814292094      0
#> 767  -0.5009217648      0
#> 768   1.0757728683      0
#> 769  -1.2005456182      0
#> 770   1.5066718538      0
#> 771   0.9088263566      0
#> 772   0.7451102473      0
#> 773   0.0279109267      0
#> 774   1.1029185439      0
#> 775  -1.1041240245      0
#> 776   0.0980697715      0
#> 777   0.2638648554      0
#> 778   1.2449192288      0
#> 779  -1.3988527975      0
#> 780  -2.2769870748      0
#> 781   0.7091367010      0
#> 782   0.2576691715      0
#> 783  -0.0775700382      0
#> 784  -0.5200717644      0
#> 785   1.6698207365      0
#> 786   2.2143037537      0
#> 787   0.5911306042      0
#> 788   1.1683624291      0
#> 789  -1.7998781518      0
#> 790   0.8398115447      0
#> 791  -0.6991771788      0
#> 792  -0.8980235958      0
#> 793  -0.8633603668      0
#> 794   0.4899071124      0
#> 795   0.4404779760      0
#> 796  -0.8600482667      0
#> 797  -0.3559888732      0
#> 798   1.7258290477      0
#> 799  -1.5389778250      0
#> 800  -0.1693716204      0
#> 801   0.0105168578      0
#> 802  -0.5686407400      0
#> 803  -0.5961889100      0
#> 804  -0.3121272322      0
#> 805   0.5331694174      0
#> 806   1.0854822294      0
#> 807  -0.0140897453      0
#> 808  -0.8507458850      0
#> 809   2.1879318330      0
#> 810  -0.0377669806      0
#> 811  -0.0629977223      0
#> 812   0.2249377587      0
#> 813  -0.4488294516      0
#> 814  -0.4376648686      0
#> 815  -0.1769919482      0
#> 816   0.0637943478      0
#> 817  -1.4754320474      0
#> 818  -1.1332202538      0
#> 819   0.3410962510      0
#> 820   1.6267735571      0
#> 821  -0.3838856236      0
#> 822  -0.6790605260      0
#> 823  -1.3830547673      0
#> 824   1.7235080997      0
#> 825   0.0154690016      0
#> 826  -1.8941319113      0
#> 827   0.6243331881      0
#> 828   1.6223988463      0
#> 829   0.6791972851      0
#> 830   0.9946508977      0
#> 831  -1.8331556390      0
#> 832  -0.1110271162      0
#> 833   1.2034782947      0
#> 834   0.4186359166      0
#> 835   0.8989642137      0
#> 836   1.3396954560      0
#> 837   0.8081903072      0
#> 838  -3.0105800531      0
#> 839   0.1735874037      0
#> 840  -1.5925640700      0
#> 841   0.1206025643      0
#> 842  -0.7207188238      0
#> 843   1.3090607466      0
#> 844   0.5456986258      0
#> 845  -0.0859142109      0
#> 846  -0.1315694763      0
#> 847  -1.5864888100      0
#> 848   1.1266767367      0
#> 849   0.8052426982      0
#> 850  -1.7826990131      0
#> 851   1.1509075033      0
#> 852   0.7260990259      0
#> 853  -1.2780444344      0
#> 854  -1.4005236372      0
#> 855   0.9980114338      0
#> 856   0.2806577933      0
#> 857   0.9699761234      0
#> 858  -0.2520465992      0
#> 859   0.3434800741      0
#> 860   0.9131648054      0
#> 861  -0.7522478516      0
#> 862  -0.3780043950      0
#> 863   0.3155040085      0
#> 864   1.2708916202      0
#> 865  -1.2763452602      0
#> 866  -0.1510476149      0
#> 867  -2.7702591148      0
#> 868   0.4598998323      0
#> 869   0.2955657275      0
#> 870   1.0319645458      0
#> 871   0.0810020875      0
#> 872  -0.9929900742      0
#> 873   0.3020559979      0
#> 874   1.6103233172      0
#> 875  -0.1484554402      0
#> 876  -1.1351666809      0
#> 877  -1.3533936636      0
#> 878   0.5909068453      0
#> 879  -0.1801607178      0
#> 880   1.0000386861      0
#> 881   0.6111684607      0
#> 882   1.9211908255      0
#> 883   0.0570911048      0
#> 884   1.6533265035      0
#> 885   1.0971415707      0
#> 886   1.3140554268      0
#> 887   0.1767843722      0
#> 888   0.0839784543      0
#> 889   0.3015840040      0
#> 890   0.7574533025      0
#> 891  -1.4933744517      0
#> 892  -0.8804923046      0
#> 893   0.8816890107      0
#> 894   1.3264187671      0
#> 895  -0.5399946304      0
#> 896  -0.5427263254      0
#> 897  -1.7412200831      0
#> 898   1.3360849932      0
#> 899  -0.2990357295      1
#> 900  -0.2807943160      1
#> 901   2.1273589198      1
#> 902   1.1156426855      1
#> 903   3.1038546743      1
#> 904  -0.1622506860      1
#> 905   3.0118652431      1
#> 906   0.5803807817      1
#> 907   1.4954083249      1
#> 908  -0.3752067757      1
#> 909   0.2199277876      1
#> 910   1.2360273514      1
#> 911  -0.1138275712      1
#> 912   1.2741273163      1
#> 913   0.0595096638      1
#> 914   1.5859582119      1
#> 915   1.1863567334      1
#> 916   1.4227481675      1
#> 917   1.0021605025      1
#> 918   0.8213678270      1
#> 919   0.7589552042      1
#> 920   2.8987862497      1
#> 921  -0.3038867653      1
#> 922   0.2572174376      1
#> 923   1.0724576826      1
#> 924   1.3108362957      1
#> 925   1.6678951927      1
#> 926   1.5550464580      1
#> 927   0.7793083455      1
#> 928   0.5892761485      1
#> 929   0.9071108901      1
#> 930   0.9982221852      1
#> 931  -0.4840739563      1
#> 932   2.2823985065      1
#> 933   0.4250543480      1
#> 934  -2.4427859520      1
#> 935   1.9363368258      1
#> 936   0.9521256425      1
#> 937   1.5631606177      1
#> 938   0.2863970808      1
#> 939   0.1298468756      1
#> 940   1.0369526329      1
#> 941   2.4667858673      1
#> 942   1.1658447542      1
#> 943  -0.2508188373      1
#> 944  -0.2540409792      1
#> 945  -0.0623980209      1
#> 946   0.9976586747      1
#> 947   0.0492928714      1
#> 948   0.2092495945      1
#> 949   1.4886386795      1
#> 950   1.6744215447      1
#> 951   0.8305549998      1
#> 952   1.8730713595      1
#> 953   0.7325970967      1
#> 954   0.4200543047      1
#> 955   1.0526248073      1
#> 956  -1.0728723117      1
#> 957   1.5167734629      1
#> 958   0.3405390821      1
#> 959   0.6010647560      1
#> 960   0.7568345849      1
#> 961   1.9588787369      1
#> 962  -1.6655283304      1
#> 963   0.6848744172      1
#> 964   0.1606107225      1
#> 965  -0.5582345267      1
#> 966   2.7919604580      1
#> 967   0.5283497696      1
#> 968   1.4944111011      1
#> 969  -0.5908977517      1
#> 970  -0.1399195295      1
#> 971  -0.3121445562      1
#> 972   1.2766870590      1
#> 973  -0.5961284297      1
#> 974   1.0062167956      1
#> 975   0.1528240264      1
#> 976   0.9964305496      1
#> 977   0.5197238172      1
#> 978  -0.1741411416      1
#> 979   0.9114119275      1
#> 980   0.3036981546      1
#> 981   2.3181815154      1
#> 982   1.2563319065      1
#> 983   0.3022892943      1
#> 984   1.6382012394      1
#> 985   0.2595121039      1
#> 986   1.4482275573      1
#> 987   0.5738869443      1
#> 988   0.2695355371      1
#> 989  -0.0572095969      1
#> 990   0.9371983886      1
#> 991   1.7435929449      1
#> 992   2.0345987953      1
#> 993   1.3258232436      1
#> 994   1.0847605581      1
#> 995   2.1023970120      1
#> 996   0.9403610787      1
#> 997   0.3067963909      1
#> 998  -0.2013688598      1
#> 999   0.9003505174      1
#> 1000 -0.0033655288      1