Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -1.292024860      0
#> 2     0.458390476      0
#> 3     0.025083172      0
#> 4     1.259447628      0
#> 5     0.803893275      0
#> 6     0.697624617      0
#> 7    -0.398727129      0
#> 8    -1.495357087      0
#> 9    -0.348744633      0
#> 10   -1.554997099      0
#> 11   -1.305412739      0
#> 12   -0.654645639      0
#> 13    0.711206375      0
#> 14    0.181960415      0
#> 15    1.464097741      0
#> 16   -0.860114609      0
#> 17   -0.417172910      0
#> 18    0.218524096      0
#> 19    1.233665318      0
#> 20   -1.895135479      0
#> 21    0.005903786      0
#> 22   -0.332313816      0
#> 23    1.382111300      0
#> 24    1.177798642      0
#> 25    0.688677548      0
#> 26    1.986287133      0
#> 27   -0.607009143      0
#> 28   -0.367341046      0
#> 29   -1.114557250      0
#> 30   -0.306326967      0
#> 31   -0.770396900      0
#> 32   -0.068369946      0
#> 33    0.515987761      0
#> 34    3.859106555      0
#> 35    2.027137461      0
#> 36    0.520930305      0
#> 37   -1.406760549      0
#> 38   -1.471625470      0
#> 39    1.530302614      0
#> 40   -0.058789242      0
#> 41    0.548925708      0
#> 42   -0.599674216      0
#> 43    0.829497332      0
#> 44    0.346472701      0
#> 45    0.175068677      0
#> 46    2.203563798      0
#> 47    0.918861378      0
#> 48    0.328779226      0
#> 49   -0.256360536      0
#> 50   -0.265448913      0
#> 51    1.631215733      0
#> 52   -0.481810160      0
#> 53    1.498879690      0
#> 54    1.082122886      0
#> 55   -0.543607590      0
#> 56    0.957593318      0
#> 57   -0.294698738      0
#> 58   -1.875460905      0
#> 59   -0.280706361      0
#> 60    1.465901111      0
#> 61   -0.234182930      0
#> 62    1.482966680      0
#> 63    0.514718253      0
#> 64   -0.516210083      0
#> 65   -2.198357902      0
#> 66    1.108316231      0
#> 67   -0.480033293      0
#> 68    0.839706001      0
#> 69    0.952211969      0
#> 70   -0.612375775      0
#> 71   -0.190167530      0
#> 72   -1.454212699      0
#> 73    1.963778082      0
#> 74   -2.389796473      0
#> 75    0.975893184      0
#> 76   -0.769575729      0
#> 77    0.751689368      0
#> 78   -0.945871020      0
#> 79   -1.230043023      0
#> 80   -0.973540611      0
#> 81    0.061603889      0
#> 82   -0.105190570      0
#> 83   -2.157852002      0
#> 84    0.411946876      0
#> 85   -0.368155552      0
#> 86    1.281486848      0
#> 87    0.465199267      0
#> 88   -0.559191393      0
#> 89   -1.919004604      0
#> 90    1.177026565      0
#> 91    0.562757345      0
#> 92   -0.069259989      0
#> 93   -0.248547134      0
#> 94    0.312235472      0
#> 95   -0.746199289      0
#> 96    1.015551663      0
#> 97    0.964689628      0
#> 98   -0.788673901      0
#> 99    0.879235657      0
#> 100   0.446437256      0
#> 101   0.140888459      0
#> 102  -1.451656590      0
#> 103   0.426328300      0
#> 104   2.487918421      0
#> 105  -0.231962617      0
#> 106  -0.889884889      0
#> 107   0.187218961      0
#> 108   0.690541008      0
#> 109   0.707174275      0
#> 110  -0.070065591      0
#> 111  -0.390970875      0
#> 112   0.457228963      0
#> 113   0.116416661      0
#> 114  -0.497121088      0
#> 115   2.037222244      0
#> 116   0.552923846      0
#> 117   1.509215569      0
#> 118   0.727422692      0
#> 119  -0.953378620      0
#> 120   0.241109312      0
#> 121   0.856284653      0
#> 122   0.740455943      0
#> 123  -0.203801428      0
#> 124   0.675738216      0
#> 125   0.230262388      0
#> 126   0.419987855      0
#> 127  -0.434993307      0
#> 128  -0.345314659      0
#> 129  -1.152465619      0
#> 130  -0.273819029      0
#> 131  -0.831590871      0
#> 132  -0.678970195      0
#> 133   0.455110962      0
#> 134   0.385536046      0
#> 135  -1.072221596      0
#> 136  -1.112124975      0
#> 137   1.595626941      0
#> 138  -0.873991919      0
#> 139  -0.793357342      0
#> 140   0.763889890      0
#> 141  -0.394466098      0
#> 142   1.248303162      0
#> 143   0.945790740      0
#> 144  -1.277879450      0
#> 145   0.679905943      0
#> 146   0.995042447      0
#> 147  -0.158598987      0
#> 148   1.070586909      0
#> 149  -0.432669839      0
#> 150   1.036075095      0
#> 151  -0.811136381      0
#> 152   1.536570202      0
#> 153   0.451511518      0
#> 154  -0.270555589      0
#> 155   0.032483117      0
#> 156  -1.057201934      0
#> 157  -2.139463308      0
#> 158  -0.588073208      0
#> 159   0.054488423      0
#> 160  -0.258323183      0
#> 161   1.023941332      0
#> 162   0.948459539      0
#> 163  -1.119908475      0
#> 164  -0.007580717      0
#> 165   0.085642403      0
#> 166  -1.056441408      0
#> 167  -1.142809695      0
#> 168  -1.124609242      0
#> 169  -0.326358798      0
#> 170   1.191337987      0
#> 171  -0.594402205      0
#> 172  -1.856043372      0
#> 173   1.354712341      0
#> 174  -0.852004970      0
#> 175  -0.006863928      0
#> 176   0.275707347      0
#> 177  -0.641867698      0
#> 178  -0.687587922      0
#> 179  -0.482009234      0
#> 180  -0.102748870      0
#> 181  -1.544774408      0
#> 182  -1.706638294      0
#> 183  -0.011932609      0
#> 184  -0.033102598      0
#> 185   1.979870025      0
#> 186   1.388067880      0
#> 187   3.038059363      0
#> 188   1.035400521      0
#> 189   0.749082027      0
#> 190   1.470616887      0
#> 191  -0.447040767      0
#> 192   0.967885430      0
#> 193  -0.896756729      0
#> 194   0.222239360      0
#> 195  -1.485259577      0
#> 196   1.264976537      0
#> 197  -1.250429864      0
#> 198  -1.597343202      0
#> 199   0.469010721      0
#> 200   0.980747177      0
#> 201   0.827031584      0
#> 202   0.444497460      0
#> 203  -0.470684474      0
#> 204   0.048210024      0
#> 205   0.789896076      0
#> 206  -0.824838157      0
#> 207  -1.318112516      0
#> 208   0.121883232      0
#> 209   2.090028325      0
#> 210   1.196663740      0
#> 211  -1.231271108      0
#> 212  -0.696647347      0
#> 213  -0.916690448      0
#> 214   0.229242136      0
#> 215  -0.812322011      0
#> 216   0.746050818      0
#> 217   0.142556511      0
#> 218  -0.397018184      0
#> 219  -1.083566222      0
#> 220   0.725109044      0
#> 221   0.173238040      0
#> 222   0.179568955      0
#> 223  -0.251623688      0
#> 224  -0.506205938      0
#> 225   0.578013707      0
#> 226   0.176537380      0
#> 227   0.909610482      0
#> 228  -0.028594280      0
#> 229   0.012900439      0
#> 230   0.811239070      0
#> 231  -0.431637687      0
#> 232   0.403672241      0
#> 233   0.211888501      0
#> 234  -0.024485474      0
#> 235  -0.401212666      0
#> 236  -0.369018750      0
#> 237   0.426256158      0
#> 238  -0.194387569      0
#> 239   0.648707687      0
#> 240  -0.217482259      0
#> 241   0.548959202      0
#> 242   0.513984397      0
#> 243  -1.925357150      0
#> 244  -0.148446650      0
#> 245  -0.378864833      0
#> 246  -1.188287133      0
#> 247  -1.575738235      0
#> 248   1.682959904      0
#> 249   0.062871248      0
#> 250  -0.244146532      0
#> 251   0.729536946      0
#> 252   2.098187897      0
#> 253   1.839563417      0
#> 254   0.913460203      0
#> 255  -1.477506174      0
#> 256   0.342657580      0
#> 257  -0.021134981      0
#> 258   1.327067920      0
#> 259   0.755460737      0
#> 260  -2.479021658      0
#> 261  -1.560449860      0
#> 262  -0.040624558      0
#> 263  -0.357629373      0
#> 264  -1.091948594      0
#> 265  -0.480763125      0
#> 266   1.013188574      0
#> 267  -1.037112591      0
#> 268   0.497204441      0
#> 269   0.105090645      0
#> 270   1.161060113      0
#> 271  -0.693192266      0
#> 272   0.892965935      0
#> 273  -0.098080420      0
#> 274   0.558912235      0
#> 275   1.587968561      0
#> 276  -0.804493058      0
#> 277   1.378507940      0
#> 278   1.938371472      0
#> 279   0.293174004      0
#> 280   0.177620376      0
#> 281  -0.745995123      0
#> 282  -0.270813727      0
#> 283   0.887763323      0
#> 284   1.042578081      0
#> 285   0.159564038      0
#> 286  -0.697309229      0
#> 287  -0.992608787      0
#> 288  -0.231936149      0
#> 289  -0.155571280      0
#> 290   0.900973546      0
#> 291  -0.125590332      0
#> 292   0.365127064      0
#> 293  -0.492435314      0
#> 294  -0.107710378      0
#> 295   0.285860911      0
#> 296   1.047317805      0
#> 297  -1.543320952      0
#> 298   0.028486390      0
#> 299   0.027733406      0
#> 300  -0.872009390      0
#> 301   0.985089756      0
#> 302   0.032542444      0
#> 303  -1.349721404      0
#> 304   0.035474995      0
#> 305   0.532915174      0
#> 306  -0.959061993      0
#> 307   0.075032684      0
#> 308  -0.385747830      0
#> 309   1.770240126      0
#> 310  -0.328366149      0
#> 311   1.371224884      0
#> 312  -0.366080925      0
#> 313   0.617230190      0
#> 314   2.161414723      0
#> 315   0.261345907      0
#> 316   0.739511654      0
#> 317  -1.897090827      0
#> 318   0.067445543      0
#> 319  -0.407551689      0
#> 320  -0.111387980      0
#> 321  -1.240766371      0
#> 322  -1.841760951      0
#> 323  -1.621832074      0
#> 324   1.202506065      0
#> 325  -1.922525034      0
#> 326   0.626570065      0
#> 327   0.232265842      0
#> 328  -1.047881376      0
#> 329   0.214812013      0
#> 330  -0.012960424      0
#> 331  -0.109147988      0
#> 332   1.578339507      0
#> 333   0.223035600      0
#> 334  -1.413835967      0
#> 335  -0.238143527      0
#> 336   1.478388065      0
#> 337   1.020946226      0
#> 338   0.933292184      0
#> 339   0.349537820      0
#> 340   0.443014446      0
#> 341  -1.619014016      0
#> 342   1.209991629      0
#> 343   0.759888276      0
#> 344   0.152908876      0
#> 345  -1.452196084      0
#> 346  -1.787771656      0
#> 347  -0.744588717      0
#> 348  -0.028832679      0
#> 349   0.082145119      0
#> 350  -1.783364835      0
#> 351   0.420186453      0
#> 352   0.660307388      0
#> 353  -0.070986549      0
#> 354   1.434107113      0
#> 355   0.009932688      0
#> 356  -1.550564362      0
#> 357   0.545646137      0
#> 358  -1.504913696      0
#> 359  -0.358279340      0
#> 360  -0.992081368      0
#> 361  -0.982365874      0
#> 362  -0.105037254      0
#> 363   1.113698899      0
#> 364  -0.616524177      0
#> 365  -0.214504272      0
#> 366  -0.496590079      0
#> 367   1.289638495      0
#> 368  -0.802614156      0
#> 369   0.493786208      0
#> 370  -1.535710926      0
#> 371   0.232533221      0
#> 372  -0.465252653      0
#> 373  -0.383896885      0
#> 374  -0.295585585      0
#> 375  -0.873436777      0
#> 376   0.986024718      0
#> 377   0.762089809      0
#> 378   1.098902491      0
#> 379   0.667576307      0
#> 380  -0.581438336      0
#> 381  -1.066054088      0
#> 382  -0.578839640      0
#> 383  -1.000838908      0
#> 384   0.927121097      0
#> 385   0.524485112      0
#> 386   0.414375079      0
#> 387   0.464129833      0
#> 388  -0.005863324      0
#> 389   0.435749287      0
#> 390   0.465658674      0
#> 391  -0.126093173      0
#> 392   1.674071069      0
#> 393  -0.836712808      0
#> 394   1.097154020      0
#> 395   0.212196054      0
#> 396   0.066410447      0
#> 397  -1.169894960      0
#> 398   0.132791329      0
#> 399  -1.073798354      0
#> 400   0.044673293      0
#> 401  -0.525313228      0
#> 402  -0.735207876      0
#> 403  -0.904173634      0
#> 404   0.179264587      0
#> 405   1.006466073      0
#> 406   2.154854986      0
#> 407  -0.907761237      0
#> 408  -0.154220716      0
#> 409   0.030276027      0
#> 410   0.275749426      0
#> 411  -0.060038133      0
#> 412  -1.495574723      0
#> 413  -1.204413187      0
#> 414   0.686963173      0
#> 415  -0.229031628      0
#> 416   0.343224163      0
#> 417   0.788256424      0
#> 418  -1.050766328      0
#> 419   1.308340228      0
#> 420  -1.318052502      0
#> 421   1.294739310      0
#> 422  -0.201685082      0
#> 423  -1.336317933      0
#> 424   0.985806568      0
#> 425   0.704257973      0
#> 426  -0.966217967      0
#> 427   0.278613579      0
#> 428   0.177537553      0
#> 429   0.583300497      0
#> 430   0.751216688      0
#> 431  -0.462640156      0
#> 432   0.281855300      0
#> 433  -1.374380111      0
#> 434   0.143398293      0
#> 435  -1.282853401      0
#> 436  -0.379395649      0
#> 437  -1.219136731      0
#> 438  -1.353913106      0
#> 439  -0.827593991      0
#> 440  -1.833495290      0
#> 441  -0.364146632      0
#> 442  -0.890943027      0
#> 443  -1.098409080      0
#> 444  -0.983080441      0
#> 445  -0.946331342      0
#> 446  -0.394296028      0
#> 447   0.033782866      0
#> 448  -0.606813449      0
#> 449  -0.103144149      0
#> 450  -0.637234683      0
#> 451   2.125754511      0
#> 452  -0.181195867      0
#> 453   0.105209141      0
#> 454   0.790663802      0
#> 455  -0.567511391      0
#> 456  -0.152743165      0
#> 457  -0.474823558      0
#> 458  -0.196318887      0
#> 459   1.167473478      0
#> 460   0.885477147      0
#> 461  -0.493877542      0
#> 462  -0.935547413      0
#> 463  -1.058823955      0
#> 464  -0.506662329      0
#> 465   1.029227663      0
#> 466  -2.337571543      0
#> 467  -2.419314810      0
#> 468  -0.479183042      0
#> 469  -0.563300158      0
#> 470   0.309950617      0
#> 471   0.537677072      0
#> 472   1.075062727      0
#> 473   0.318066710      0
#> 474   1.432791229      0
#> 475  -0.941664519      0
#> 476   0.270427206      0
#> 477   0.506400947      0
#> 478   1.496858195      0
#> 479  -2.417904168      0
#> 480  -2.582883790      0
#> 481   1.612530589      0
#> 482  -0.914969551      0
#> 483  -1.180796664      0
#> 484  -0.415526130      0
#> 485  -0.657390572      0
#> 486   1.362553031      0
#> 487   0.002138021      0
#> 488   0.987593390      0
#> 489   0.038123650      0
#> 490  -0.094084364      0
#> 491  -0.083106397      0
#> 492  -1.658208107      0
#> 493   1.770255831      0
#> 494  -0.996622834      0
#> 495   0.014129308      0
#> 496  -0.858654351      0
#> 497   0.707152426      0
#> 498   0.742248673      0
#> 499   0.542219044      0
#> 500   1.390361818      0
#> 501   0.772871242      0
#> 502   0.669545934      0
#> 503  -0.209066792      0
#> 504  -0.566913409      0
#> 505  -1.228919218      0
#> 506  -1.569880640      0
#> 507   0.653460075      0
#> 508   0.066430568      0
#> 509  -1.778943739      0
#> 510   0.262803385      0
#> 511   0.688153496      0
#> 512  -0.393460375      0
#> 513   1.635899315      0
#> 514   0.419557106      0
#> 515  -1.803922253      0
#> 516   0.448484934      0
#> 517   0.274853385      0
#> 518   1.759622665      0
#> 519  -0.264076396      0
#> 520   0.959555857      0
#> 521  -1.107802670      0
#> 522   0.942208415      0
#> 523   0.072951909      0
#> 524  -0.443584114      0
#> 525  -0.390898662      0
#> 526   2.610672616      0
#> 527   1.213049547      0
#> 528   0.280896186      0
#> 529   1.519391935      0
#> 530  -0.279957389      0
#> 531  -0.500906084      0
#> 532   0.330527479      0
#> 533  -0.032268045      0
#> 534  -1.142975217      0
#> 535   1.357852121      0
#> 536   0.317510112      0
#> 537   0.151171230      0
#> 538  -1.562287181      0
#> 539   0.576362472      0
#> 540  -1.957211742      0
#> 541   0.633615676      0
#> 542  -0.911212457      0
#> 543   1.418329940      0
#> 544   0.801644219      0
#> 545   1.653254529      0
#> 546   0.316876112      0
#> 547   0.268624016      0
#> 548  -0.108614678      0
#> 549   0.738399157      0
#> 550  -0.059581642      0
#> 551   0.877503712      0
#> 552  -0.182889965      0
#> 553   0.736142678      0
#> 554   1.180703220      0
#> 555  -0.873761645      0
#> 556   1.330905629      0
#> 557  -0.230782331      0
#> 558  -1.163457275      0
#> 559   0.516519455      0
#> 560   1.348799687      0
#> 561   0.781019926      0
#> 562  -1.336368028      0
#> 563   1.708977134      0
#> 564  -2.421047558      0
#> 565   1.652592293      0
#> 566   0.622690211      0
#> 567   0.167134800      0
#> 568   0.724922128      0
#> 569   0.292131534      0
#> 570  -0.353662145      0
#> 571   1.464508846      0
#> 572  -0.796329040      0
#> 573  -0.419585565      0
#> 574   0.092990526      0
#> 575   0.168586479      0
#> 576  -0.296528320      0
#> 577  -0.367145929      0
#> 578  -0.797062286      0
#> 579  -0.309576399      0
#> 580  -1.299776289      0
#> 581   0.543592362      0
#> 582   1.678259925      0
#> 583   1.204940222      0
#> 584   0.887223405      0
#> 585  -0.078173939      0
#> 586   1.636944383      0
#> 587  -1.188771593      0
#> 588   0.302093121      0
#> 589   1.636046926      0
#> 590   0.098427855      0
#> 591   0.358198190      0
#> 592   0.665329665      0
#> 593   0.751271166      0
#> 594   1.763593684      0
#> 595  -1.482433801      0
#> 596  -0.083630604      0
#> 597   0.194805968      0
#> 598   0.437985419      0
#> 599  -0.217132124      0
#> 600  -0.846431780      0
#> 601   0.353878847      0
#> 602   0.121042180      0
#> 603   0.006899922      0
#> 604   0.084174403      0
#> 605   0.126588442      0
#> 606  -0.717852397      0
#> 607  -1.253208172      0
#> 608   2.044024049      0
#> 609  -0.065238137      0
#> 610  -0.152796094      0
#> 611  -0.131340101      0
#> 612   0.605998613      0
#> 613   0.613927557      0
#> 614  -1.966551971      0
#> 615  -0.019029870      0
#> 616   0.537040656      0
#> 617  -0.077128789      0
#> 618  -2.841663093      0
#> 619   0.467556065      0
#> 620   0.194205275      0
#> 621  -0.115489158      0
#> 622   0.165342352      0
#> 623  -0.391679092      0
#> 624   1.153495983      0
#> 625  -0.224770682      0
#> 626   0.878867430      0
#> 627   0.387146176      0
#> 628  -1.752462717      0
#> 629   2.444452785      0
#> 630  -0.980236804      0
#> 631   1.971722975      0
#> 632  -1.004414730      0
#> 633   0.528983387      0
#> 634  -1.601363726      0
#> 635  -0.726464300      0
#> 636  -0.615962263      0
#> 637  -1.429456187      0
#> 638  -0.561142958      0
#> 639   1.819363505      0
#> 640   1.447994958      0
#> 641   0.205154203      0
#> 642  -0.095262707      0
#> 643  -1.059706184      0
#> 644  -1.340101610      0
#> 645  -0.679678477      0
#> 646   0.628476587      0
#> 647  -0.516836048      0
#> 648  -1.565420236      0
#> 649   0.152334755      0
#> 650   0.741553214      0
#> 651  -0.858845703      0
#> 652   0.713313859      0
#> 653  -0.211744338      0
#> 654  -0.358723436      0
#> 655  -0.541330931      0
#> 656   0.645636000      0
#> 657   0.902606988      0
#> 658  -0.236653652      0
#> 659  -0.288864908      0
#> 660   0.175247198      0
#> 661  -0.755382382      0
#> 662   0.445850950      0
#> 663   0.556353719      0
#> 664   1.539606756      0
#> 665   1.106441951      0
#> 666   0.331427709      0
#> 667  -0.566581214      0
#> 668   1.022223974      0
#> 669   1.453302190      0
#> 670   0.959803351      0
#> 671   0.060102946      0
#> 672   0.928039352      0
#> 673   0.947123738      0
#> 674   0.231292722      0
#> 675  -0.352516586      0
#> 676  -0.317048992      0
#> 677   0.177884075      0
#> 678  -0.462722981      0
#> 679   0.332156081      0
#> 680   1.586261563      0
#> 681   1.734331400      0
#> 682  -1.328465411      0
#> 683   1.768784279      0
#> 684  -0.942588383      0
#> 685   0.028007477      0
#> 686   0.026362662      0
#> 687  -0.081123557      0
#> 688   0.107163910      0
#> 689   1.162829437      0
#> 690   0.310966385      0
#> 691   0.503423111      0
#> 692   0.602512136      0
#> 693   0.358646320      0
#> 694  -0.193487642      0
#> 695  -0.199181913      0
#> 696   0.043804016      0
#> 697  -0.046124540      0
#> 698   0.334512031      0
#> 699  -3.026433653      0
#> 700   0.837049387      0
#> 701  -1.298610166      0
#> 702  -0.025933464      0
#> 703  -0.090320560      0
#> 704  -0.401826480      0
#> 705   0.515781266      0
#> 706  -0.625159803      0
#> 707  -1.272368497      0
#> 708  -1.545784896      0
#> 709  -1.705157280      0
#> 710   0.794236582      0
#> 711  -0.020927271      0
#> 712  -0.629062586      0
#> 713  -0.312561280      0
#> 714   0.178671098      0
#> 715   0.489169688      0
#> 716  -0.908990142      0
#> 717   0.872698615      0
#> 718   1.139053859      0
#> 719   0.373414863      0
#> 720  -0.394911177      0
#> 721  -2.123000530      0
#> 722  -1.272299468      0
#> 723  -0.555373393      0
#> 724   1.556105136      0
#> 725  -1.099329476      0
#> 726   1.906710261      0
#> 727   0.514868882      0
#> 728  -0.681129717      0
#> 729  -0.105760814      0
#> 730  -0.969586800      0
#> 731  -0.075562726      0
#> 732   0.665663640      0
#> 733  -1.851652034      0
#> 734  -1.429388635      0
#> 735   0.853476724      0
#> 736   2.427330850      0
#> 737  -1.141047044      0
#> 738  -2.518068856      0
#> 739  -0.128222613      0
#> 740  -1.647711816      0
#> 741   0.913351717      0
#> 742   0.259277499      0
#> 743   0.635888079      0
#> 744  -0.876118624      0
#> 745   0.537195676      0
#> 746  -0.112560366      0
#> 747  -0.579629778      0
#> 748   0.028217887      0
#> 749  -1.714509943      0
#> 750   0.843177791      0
#> 751  -0.259719289      0
#> 752  -0.270896268      0
#> 753  -0.959474657      0
#> 754  -0.653807226      0
#> 755   1.175850330      0
#> 756  -0.882667218      0
#> 757  -0.811790863      0
#> 758   2.149409384      0
#> 759  -0.383595624      0
#> 760   0.557059678      0
#> 761   0.003388191      0
#> 762  -0.007816947      0
#> 763   0.677007673      0
#> 764  -0.160497216      0
#> 765   1.524885817      0
#> 766  -0.607855369      0
#> 767   1.109096934      0
#> 768   1.209086901      0
#> 769  -0.253433783      0
#> 770  -0.682780836      0
#> 771   2.267281435      0
#> 772  -0.926091338      0
#> 773  -1.107681637      0
#> 774  -2.009793871      0
#> 775  -1.334049643      0
#> 776   0.450018809      0
#> 777  -1.559830975      0
#> 778  -0.490669876      0
#> 779  -2.231448369      0
#> 780  -0.019946111      0
#> 781  -0.803287125      0
#> 782   0.001664708      0
#> 783  -0.943563187      0
#> 784  -0.209434258      0
#> 785  -0.881546401      0
#> 786   0.594686117      0
#> 787   0.672787480      0
#> 788  -0.817691520      0
#> 789  -1.555059025      0
#> 790   0.173124430      0
#> 791   1.426698521      0
#> 792   0.367205162      0
#> 793  -1.193251186      0
#> 794  -1.318013947      0
#> 795  -0.165347819      0
#> 796   0.607787003      0
#> 797   0.354430765      0
#> 798  -2.325495218      0
#> 799   0.823044275      0
#> 800   0.569867715      0
#> 801   0.761163035      0
#> 802  -1.461193335      0
#> 803  -0.112842560      0
#> 804  -1.002250194      0
#> 805  -0.032512312      0
#> 806   0.203173083      0
#> 807   1.572969350      0
#> 808  -0.401798277      0
#> 809  -0.993321611      0
#> 810   0.286925359      0
#> 811   1.666889804      0
#> 812  -0.589381906      0
#> 813   0.076543984      0
#> 814  -1.211794185      0
#> 815  -0.818891188      0
#> 816  -1.380805075      0
#> 817   0.612967516      0
#> 818  -0.466136208      0
#> 819   0.958374124      0
#> 820   1.305844956      0
#> 821   0.652413700      0
#> 822   0.721707330      0
#> 823   0.644748504      0
#> 824   0.963067003      0
#> 825   0.246581120      0
#> 826   1.310193804      0
#> 827  -0.936940332      0
#> 828   0.642345983      0
#> 829  -0.973688208      0
#> 830   1.163669630      0
#> 831  -0.296527352      0
#> 832   0.597357762      0
#> 833  -0.271172706      0
#> 834  -1.093563299      0
#> 835  -1.295027401      0
#> 836   1.560713748      0
#> 837  -0.444112356      0
#> 838   0.595885313      0
#> 839  -0.019124401      0
#> 840  -0.990709498      0
#> 841  -0.057868774      0
#> 842  -0.236371700      0
#> 843  -0.562831222      0
#> 844   0.126661245      0
#> 845  -1.253447822      0
#> 846  -0.604582224      0
#> 847   0.464101253      0
#> 848  -2.021674581      0
#> 849  -0.949218373      0
#> 850  -0.490393734      0
#> 851  -0.973412016      0
#> 852   0.773595230      0
#> 853  -0.172136611      0
#> 854  -1.303278816      0
#> 855   0.855661137      0
#> 856  -0.638705704      0
#> 857   0.005523948      0
#> 858  -0.061977282      0
#> 859  -0.404443127      0
#> 860   0.680820112      0
#> 861   0.680440739      0
#> 862  -0.899170016      0
#> 863  -1.793535932      0
#> 864  -0.837427850      0
#> 865   1.911375985      0
#> 866  -0.069103915      0
#> 867  -0.608542338      0
#> 868   0.639348248      0
#> 869   0.043011281      0
#> 870   0.621264998      0
#> 871   0.160269311      0
#> 872   0.479075206      0
#> 873   1.790976449      0
#> 874   1.135542785      0
#> 875   0.603256236      0
#> 876   1.471609486      0
#> 877   0.013060828      0
#> 878   1.386606398      0
#> 879   0.030925733      0
#> 880  -0.148580627      0
#> 881  -0.113572612      0
#> 882   0.477915917      0
#> 883  -0.078046022      0
#> 884   0.302746981      0
#> 885  -1.320091562      0
#> 886   0.804279442      0
#> 887   1.084605321      0
#> 888   0.068122813      1
#> 889   1.883888989      1
#> 890  -0.060599027      1
#> 891   0.596798145      1
#> 892  -0.598209699      1
#> 893   0.651702627      1
#> 894   0.251550180      1
#> 895   0.873006944      1
#> 896   0.036663110      1
#> 897   1.700277993      1
#> 898   0.526603879      1
#> 899   1.166189411      1
#> 900   1.212256427      1
#> 901  -0.386341954      1
#> 902   0.550181161      1
#> 903   1.547444369      1
#> 904  -0.526929300      1
#> 905  -0.190392982      1
#> 906  -0.006272655      1
#> 907   1.812765108      1
#> 908   0.601309159      1
#> 909   1.348144724      1
#> 910   0.768942076      1
#> 911   1.961165146      1
#> 912  -0.087713521      1
#> 913   2.133659429      1
#> 914  -0.115115689      1
#> 915   1.491891770      1
#> 916   1.234521557      1
#> 917   1.807138727      1
#> 918  -0.455292370      1
#> 919   0.126415843      1
#> 920   1.283955842      1
#> 921   0.585366450      1
#> 922   2.320226770      1
#> 923  -0.387079986      1
#> 924   1.880484330      1
#> 925   1.412061007      1
#> 926   1.954468314      1
#> 927   0.449840589      1
#> 928   1.058531595      1
#> 929  -0.271300217      1
#> 930  -0.387746652      1
#> 931   1.051971336      1
#> 932   0.843040280      1
#> 933  -0.587192013      1
#> 934   1.396615016      1
#> 935  -0.363844970      1
#> 936   2.334690588      1
#> 937   0.239113781      1
#> 938   1.846505362      1
#> 939   2.348057067      1
#> 940   1.247827945      1
#> 941   4.450738055      1
#> 942   0.165583137      1
#> 943   0.508287939      1
#> 944  -1.045413969      1
#> 945  -0.177154045      1
#> 946   0.394670644      1
#> 947   0.589746597      1
#> 948   0.037610350      1
#> 949  -0.211548068      1
#> 950   0.531153146      1
#> 951   0.541573138      1
#> 952  -0.706277771      1
#> 953   2.216913848      1
#> 954   0.376643418      1
#> 955   0.863091103      1
#> 956  -0.285870275      1
#> 957   0.534156026      1
#> 958   1.001926273      1
#> 959   1.787700583      1
#> 960   1.059255919      1
#> 961   0.163550631      1
#> 962   1.362096052      1
#> 963   0.038548855      1
#> 964   1.441383045      1
#> 965   2.137410870      1
#> 966   0.319837563      1
#> 967   1.770722569      1
#> 968   2.276153963      1
#> 969  -1.067669044      1
#> 970   0.062146565      1
#> 971   0.991026670      1
#> 972   0.254591866      1
#> 973  -0.196418655      1
#> 974   0.928553517      1
#> 975   2.092459822      1
#> 976   0.698267836      1
#> 977   1.697788209      1
#> 978   0.324689423      1
#> 979  -1.644986869      1
#> 980   1.159163271      1
#> 981   0.092151428      1
#> 982   0.448978316      1
#> 983  -1.138080288      1
#> 984  -0.360165142      1
#> 985   1.043184219      1
#> 986   0.670561777      1
#> 987   1.507268269      1
#> 988   0.886927359      1
#> 989  -0.926364930      1
#> 990   2.565100614      1
#> 991   0.837424242      1
#> 992   0.028411359      1
#> 993   0.928724924      1
#> 994  -0.550061272      1
#> 995   1.669896665      1
#> 996  -0.643855588      1
#> 997  -0.622179395      1
#> 998  -0.724016232      1
#> 999   1.287245251      1
#> 1000  0.802472335      1