Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.702292951      0
#> 2    -0.382069719      0
#> 3    -0.490146166      0
#> 4    -0.588039062      0
#> 5     1.145938675      0
#> 6     0.542054611      0
#> 7    -1.288197802      0
#> 8     1.410738125      0
#> 9     1.798002624      0
#> 10   -1.594218014      0
#> 11    2.376824632      0
#> 12   -0.187462564      0
#> 13   -1.011742921      0
#> 14    3.096433187      0
#> 15    0.420727525      0
#> 16    1.452982495      0
#> 17    1.315481703      0
#> 18   -0.672823973      0
#> 19   -0.250383280      0
#> 20   -0.923749560      0
#> 21   -0.778547410      0
#> 22   -0.161035791      0
#> 23    0.212895937      0
#> 24    0.745150385      0
#> 25   -0.904418657      0
#> 26    0.429272101      0
#> 27   -1.233130279      0
#> 28   -1.299373992      0
#> 29    0.469433651      0
#> 30   -1.355329170      0
#> 31   -0.272692615      0
#> 32    0.147240986      0
#> 33    0.313306876      0
#> 34    1.719597121      0
#> 35    0.585492964      0
#> 36   -0.997420805      0
#> 37    0.191637313      0
#> 38    0.498773379      0
#> 39   -2.224895104      0
#> 40   -0.704080586      0
#> 41   -1.199019627      0
#> 42   -1.566271694      0
#> 43    0.127219411      0
#> 44    0.924277341      0
#> 45   -0.299652670      0
#> 46    1.345656946      0
#> 47    0.985557734      0
#> 48   -0.190294467      0
#> 49    0.230525817      0
#> 50    0.100228391      0
#> 51   -2.705916255      0
#> 52   -1.202569095      0
#> 53    0.536457554      0
#> 54   -2.208382791      0
#> 55   -0.300520374      0
#> 56   -0.186869348      0
#> 57    0.802117364      0
#> 58   -0.225106880      0
#> 59   -1.619574189      0
#> 60    0.324860772      0
#> 61    0.146722947      0
#> 62   -0.356548728      0
#> 63   -0.439281512      0
#> 64    0.481392339      0
#> 65    0.133707792      0
#> 66   -1.273086212      0
#> 67    2.030649018      0
#> 68   -0.222344161      0
#> 69   -0.106006998      0
#> 70   -0.395064392      0
#> 71   -0.740056117      0
#> 72   -0.581667031      0
#> 73    0.050775932      0
#> 74    0.479715352      0
#> 75   -2.035172970      0
#> 76    0.513421478      0
#> 77    0.763056099      0
#> 78    0.880742003      0
#> 79   -0.913567299      0
#> 80   -0.485261910      0
#> 81   -0.218035547      0
#> 82   -0.529345236      0
#> 83   -0.477772016      0
#> 84   -1.567280433      0
#> 85   -1.588191119      0
#> 86   -0.742751255      0
#> 87    0.725984036      0
#> 88   -2.206082259      0
#> 89    0.013124386      0
#> 90    1.608800238      0
#> 91    0.685056910      0
#> 92    1.591910153      0
#> 93    0.536811453      0
#> 94   -1.862893183      0
#> 95    0.581511867      0
#> 96    0.787871811      0
#> 97   -0.056420431      0
#> 98    1.383530248      0
#> 99    1.986840648      0
#> 100  -0.338851934      0
#> 101  -0.061964175      0
#> 102   0.052237352      0
#> 103   0.996468362      0
#> 104  -1.383401171      0
#> 105  -0.712457758      0
#> 106   0.346422517      0
#> 107   0.569951989      0
#> 108  -0.793332429      0
#> 109  -0.526820557      0
#> 110   1.813944647      0
#> 111  -0.444555218      0
#> 112   0.157979708      0
#> 113  -3.145640468      0
#> 114  -0.489119129      0
#> 115  -0.406693807      0
#> 116  -0.819959335      0
#> 117  -2.594919645      0
#> 118  -0.456274872      0
#> 119   0.210398661      0
#> 120   0.399792997      0
#> 121  -0.818538992      0
#> 122  -0.150459522      0
#> 123  -0.362009418      0
#> 124  -2.141262898      0
#> 125   0.580821088      0
#> 126  -0.780612772      0
#> 127  -1.176771891      0
#> 128   0.633939672      0
#> 129  -0.899539057      0
#> 130  -1.333181707      0
#> 131   0.479606034      0
#> 132   0.415790777      0
#> 133   0.422303653      0
#> 134   1.243321938      0
#> 135   0.975129146      0
#> 136  -0.713105974      0
#> 137   0.944328475      0
#> 138  -0.819928566      0
#> 139  -1.737649267      0
#> 140  -0.682400905      0
#> 141   0.489869932      0
#> 142   1.281742115      0
#> 143   0.586810513      0
#> 144   0.236674915      0
#> 145  -0.027828271      0
#> 146  -0.574450739      0
#> 147   0.725080729      0
#> 148   0.501997831      0
#> 149  -0.309692830      0
#> 150   0.286038930      0
#> 151   0.752136947      0
#> 152   2.421772761      0
#> 153  -0.509726702      0
#> 154  -1.187426022      0
#> 155  -0.760314095      0
#> 156   1.353847440      0
#> 157   0.092837811      0
#> 158   0.290822740      0
#> 159  -1.184628686      0
#> 160  -0.923908932      0
#> 161   0.336238925      0
#> 162  -1.192405954      0
#> 163  -2.339855384      0
#> 164   0.798510987      0
#> 165  -1.255520261      0
#> 166  -0.152937133      0
#> 167  -0.002898232      0
#> 168  -0.399563559      0
#> 169   0.970233427      0
#> 170  -0.761531488      0
#> 171  -0.918790093      0
#> 172   0.785226160      0
#> 173  -0.691786378      0
#> 174   1.328104252      0
#> 175  -1.310103283      0
#> 176   0.731058663      0
#> 177  -1.798790251      0
#> 178  -0.439024249      0
#> 179  -0.839218006      0
#> 180  -0.566810764      0
#> 181   1.549704358      0
#> 182   0.320024982      0
#> 183  -0.998335287      0
#> 184   1.347360570      0
#> 185  -0.466631070      0
#> 186   1.924772469      0
#> 187   0.922884681      0
#> 188  -0.396993288      0
#> 189   0.098777185      0
#> 190  -0.066886894      0
#> 191   1.183960708      0
#> 192   0.962483885      0
#> 193   0.346391304      0
#> 194   0.863758394      0
#> 195   1.715488068      0
#> 196   0.467120708      0
#> 197  -0.830644335      0
#> 198  -1.496976961      0
#> 199  -0.609162488      0
#> 200  -0.954001608      0
#> 201  -0.837488080      0
#> 202   0.017108509      0
#> 203  -0.765021052      0
#> 204  -0.280389092      0
#> 205   0.527723658      0
#> 206  -0.048848467      0
#> 207   1.006529105      0
#> 208  -0.538480335      0
#> 209  -0.198678216      0
#> 210   0.482788292      0
#> 211   1.135458782      0
#> 212   0.804359479      0
#> 213   1.067345867      0
#> 214  -0.871624940      0
#> 215  -0.635396554      0
#> 216  -0.156229460      0
#> 217  -0.693375811      0
#> 218   1.161282622      0
#> 219   0.449590996      0
#> 220  -0.803748283      0
#> 221   1.181641439      0
#> 222   1.297823487      0
#> 223  -0.045691435      0
#> 224  -0.591431876      0
#> 225   2.236805991      0
#> 226   0.306389061      0
#> 227   0.682487503      0
#> 228  -0.171231682      0
#> 229   2.711895016      0
#> 230  -1.601424064      0
#> 231  -0.551874425      0
#> 232   0.985779610      0
#> 233   0.930143497      0
#> 234  -1.942425433      0
#> 235  -0.299898838      0
#> 236  -0.354091945      0
#> 237  -0.721149734      0
#> 238  -1.660806140      0
#> 239  -1.034210855      0
#> 240  -0.222605895      0
#> 241  -0.170525836      0
#> 242  -0.828802807      0
#> 243   1.344885766      0
#> 244  -0.919806927      0
#> 245   1.560151411      0
#> 246  -0.471669993      0
#> 247  -1.789644585      0
#> 248  -0.943085083      0
#> 249  -1.407686993      0
#> 250   0.531409691      0
#> 251  -0.402895618      0
#> 252  -0.421527051      0
#> 253  -0.679202394      0
#> 254   0.423924790      0
#> 255  -0.655225542      0
#> 256  -0.668000339      0
#> 257   0.086722015      0
#> 258  -0.151759041      0
#> 259   0.155112177      0
#> 260  -0.191914569      0
#> 261  -1.091779475      0
#> 262  -0.225263917      0
#> 263  -0.717845537      0
#> 264  -0.575591894      0
#> 265  -0.063818889      0
#> 266  -0.605103442      0
#> 267  -2.070234585      0
#> 268   0.940215464      0
#> 269  -1.627140880      0
#> 270   1.159083501      0
#> 271   0.098429813      0
#> 272  -0.073915976      0
#> 273   0.164471845      0
#> 274   0.910133890      0
#> 275  -0.921936729      0
#> 276   0.423707270      0
#> 277   0.086505249      0
#> 278  -2.033130217      0
#> 279   0.479589074      0
#> 280   1.150589359      0
#> 281   0.236878070      0
#> 282  -0.174845507      0
#> 283  -0.748832684      0
#> 284   0.078821556      0
#> 285   2.257969919      0
#> 286  -0.233949860      0
#> 287   1.587237007      0
#> 288   1.598456467      0
#> 289   0.263839999      0
#> 290   0.078363771      0
#> 291  -0.923425519      0
#> 292  -0.893319890      0
#> 293   2.194392154      0
#> 294   0.616328825      0
#> 295  -1.165570635      0
#> 296  -0.014507523      0
#> 297   1.015235612      0
#> 298   1.988890119      0
#> 299  -1.917576495      0
#> 300   0.790704632      0
#> 301  -0.438754396      0
#> 302  -0.091887897      0
#> 303  -0.780195668      0
#> 304   0.915916133      0
#> 305   0.066008131      0
#> 306  -0.289409596      0
#> 307  -0.087149588      0
#> 308   0.044651243      0
#> 309   0.891285122      0
#> 310  -1.102776459      0
#> 311  -0.197247503      0
#> 312  -0.457443552      0
#> 313  -0.470648664      0
#> 314  -0.116493575      0
#> 315  -0.221613512      0
#> 316  -0.742613247      0
#> 317  -0.720790977      0
#> 318  -0.159843404      0
#> 319  -0.289457442      0
#> 320  -0.202228376      0
#> 321  -0.825509413      0
#> 322   0.890244888      0
#> 323  -0.214000868      0
#> 324   0.015580109      0
#> 325  -0.153880798      0
#> 326   0.815516387      0
#> 327   0.515802209      0
#> 328   1.611599839      0
#> 329  -1.160272168      0
#> 330  -0.449477178      0
#> 331  -0.192304342      0
#> 332  -0.349735147      0
#> 333  -0.592427890      0
#> 334  -1.944001697      0
#> 335  -0.525606945      0
#> 336   0.343672980      0
#> 337  -0.041279813      0
#> 338   0.064440467      0
#> 339   0.462845411      0
#> 340  -1.613562259      0
#> 341  -1.076674945      0
#> 342   1.239993758      0
#> 343   0.417059918      0
#> 344  -0.547529378      0
#> 345   0.110700936      0
#> 346  -0.239857260      0
#> 347  -0.142817439      0
#> 348   0.718403027      0
#> 349  -1.828359779      0
#> 350  -1.080267569      0
#> 351  -1.037555599      0
#> 352   0.259192656      0
#> 353  -1.415596083      0
#> 354  -0.140533314      0
#> 355  -1.979555953      0
#> 356   0.537274550      0
#> 357  -0.065714764      0
#> 358  -1.271684862      0
#> 359  -0.104738534      0
#> 360   0.752599759      0
#> 361   1.876324064      0
#> 362  -1.620517099      0
#> 363  -0.174614608      0
#> 364  -0.551593501      0
#> 365  -0.051176887      0
#> 366   0.926582488      0
#> 367  -2.104528102      0
#> 368   0.496238336      0
#> 369   0.795797819      0
#> 370  -1.832643245      0
#> 371  -0.361721802      0
#> 372  -0.315656379      0
#> 373   0.149188395      0
#> 374   0.818720828      0
#> 375   1.524064595      0
#> 376  -1.312466860      0
#> 377   1.563743767      0
#> 378  -1.240160046      0
#> 379  -0.510691633      0
#> 380  -0.526152198      0
#> 381  -0.765372489      0
#> 382   1.314150570      0
#> 383  -0.578429419      0
#> 384   1.500053386      0
#> 385   1.014317624      0
#> 386  -0.019490297      0
#> 387   1.244899986      0
#> 388  -1.736499622      0
#> 389  -1.012058483      0
#> 390   0.064216265      0
#> 391  -0.351320285      0
#> 392   2.233641078      0
#> 393  -0.699647271      0
#> 394   0.548576418      0
#> 395  -2.414518426      0
#> 396   1.134357234      0
#> 397   0.097259774      0
#> 398   0.211872340      0
#> 399   0.256967203      0
#> 400  -1.494472074      0
#> 401   0.181933077      0
#> 402   1.886685984      0
#> 403  -0.762259814      0
#> 404  -0.526646151      0
#> 405  -1.477193381      0
#> 406   1.921421626      0
#> 407   0.720863944      0
#> 408  -0.194101612      0
#> 409   1.126584774      0
#> 410  -1.356824302      0
#> 411  -0.097464971      0
#> 412  -1.429207326      0
#> 413   1.186277473      0
#> 414   0.565230065      0
#> 415  -0.046321005      0
#> 416   0.464991595      0
#> 417  -1.276295778      0
#> 418   0.541768310      0
#> 419  -0.187635826      0
#> 420  -0.884450208      0
#> 421   0.396553285      0
#> 422  -1.140245776      0
#> 423   1.557140772      0
#> 424   1.962021508      0
#> 425   0.658522088      0
#> 426   0.806422318      0
#> 427   0.308873566      0
#> 428  -0.244444097      0
#> 429  -0.430004643      0
#> 430  -1.808052793      0
#> 431  -1.124200503      0
#> 432   1.345023807      0
#> 433   1.819007775      0
#> 434   0.165149539      0
#> 435  -1.539477757      0
#> 436  -0.060475084      0
#> 437   0.062604180      0
#> 438   0.424340387      0
#> 439   1.102675551      0
#> 440   1.829520118      0
#> 441   0.671031050      0
#> 442   0.774026548      0
#> 443   0.153012275      0
#> 444   0.229107696      0
#> 445  -0.739567612      0
#> 446  -0.115702001      0
#> 447  -0.340771398      0
#> 448  -0.903820000      0
#> 449  -0.950972176      0
#> 450   0.884887041      0
#> 451  -0.147569718      0
#> 452  -0.533846988      0
#> 453  -0.484520541      0
#> 454  -0.463645361      0
#> 455  -0.704015462      0
#> 456   0.278160945      0
#> 457   0.915528316      0
#> 458   0.643414339      0
#> 459   0.599731763      0
#> 460  -1.670279065      0
#> 461   0.778081906      0
#> 462  -1.161394195      0
#> 463  -0.581699220      0
#> 464  -0.514926403      0
#> 465   0.086942417      0
#> 466  -0.805139183      0
#> 467  -0.373003242      0
#> 468  -0.278811299      0
#> 469   0.643062196      0
#> 470  -1.167984805      0
#> 471   2.991586841      0
#> 472   1.792669164      0
#> 473   1.396579503      0
#> 474  -1.908374853      0
#> 475   0.685341562      0
#> 476   0.872810300      0
#> 477  -0.726142097      0
#> 478   1.166284028      0
#> 479  -0.722646426      0
#> 480   0.427848967      0
#> 481  -0.269560897      0
#> 482   0.228937075      0
#> 483   1.642280894      0
#> 484   0.607577034      0
#> 485   0.243294862      0
#> 486  -0.617747740      0
#> 487   0.859724823      0
#> 488  -0.814348135      0
#> 489  -0.915737131      0
#> 490   0.264150081      0
#> 491   0.410206412      0
#> 492  -0.235568871      0
#> 493   1.492358655      0
#> 494  -1.342882820      0
#> 495   0.511076981      0
#> 496   0.059796766      0
#> 497  -0.995712751      0
#> 498  -0.008668213      0
#> 499  -0.119197160      0
#> 500   0.579739169      0
#> 501  -0.363635174      0
#> 502  -0.987097488      0
#> 503  -0.089242232      0
#> 504   1.506430571      0
#> 505  -1.369433872      0
#> 506   0.028005284      0
#> 507  -0.947667733      0
#> 508   0.651846251      0
#> 509   0.234799643      0
#> 510  -0.040706580      0
#> 511  -0.580001064      0
#> 512  -0.201540069      0
#> 513   1.119009870      0
#> 514  -0.519972311      0
#> 515  -0.914191097      0
#> 516  -0.336848263      0
#> 517   0.949634467      0
#> 518   1.748785434      0
#> 519   1.191286395      0
#> 520  -0.197615042      0
#> 521  -0.893956949      0
#> 522  -1.323726691      0
#> 523  -1.154766312      0
#> 524  -0.325809906      0
#> 525  -0.991107099      0
#> 526  -0.632786802      0
#> 527   0.431168661      0
#> 528  -0.261891366      0
#> 529   1.605222475      0
#> 530  -0.277716496      0
#> 531  -0.552255489      0
#> 532   0.166400074      0
#> 533   2.927033387      0
#> 534  -1.564961226      0
#> 535   0.310860363      0
#> 536   0.672761394      0
#> 537  -0.585351810      0
#> 538  -0.871812922      0
#> 539   0.271650658      0
#> 540  -0.349361001      0
#> 541  -0.612580093      0
#> 542  -1.292115264      0
#> 543   0.284181288      0
#> 544  -1.351698306      0
#> 545   1.541115120      0
#> 546  -1.500866678      0
#> 547  -1.589622561      0
#> 548   0.573633091      0
#> 549  -0.877131492      0
#> 550  -0.869779737      0
#> 551   0.894972363      0
#> 552   0.369505780      0
#> 553  -0.080305503      0
#> 554   0.262398069      0
#> 555  -0.328381146      0
#> 556   0.535108887      0
#> 557  -1.135239621      0
#> 558  -0.348851609      0
#> 559  -1.841242090      0
#> 560   0.599507864      0
#> 561  -0.871217518      0
#> 562   0.042707425      0
#> 563  -0.010845208      0
#> 564   0.113087502      0
#> 565   1.299758079      0
#> 566  -1.254961987      0
#> 567  -0.690443871      0
#> 568   1.279274425      0
#> 569  -1.609047455      0
#> 570   0.053170225      0
#> 571   0.716636425      0
#> 572  -1.642879846      0
#> 573   0.691982152      0
#> 574   0.460626904      0
#> 575  -1.453849262      0
#> 576  -0.232399274      0
#> 577  -0.395954176      0
#> 578   1.300868392      0
#> 579  -0.401239665      0
#> 580  -0.053200151      0
#> 581  -0.553180022      0
#> 582   0.971874737      0
#> 583   0.945817302      0
#> 584  -0.196698284      0
#> 585   0.057398898      0
#> 586   1.374639856      0
#> 587   0.343656340      0
#> 588   0.308420342      0
#> 589   1.295987310      0
#> 590   0.458472905      0
#> 591  -0.839183293      0
#> 592  -0.725911660      0
#> 593  -1.216173686      0
#> 594   0.387472230      0
#> 595  -1.431368776      0
#> 596  -0.066340640      0
#> 597  -0.121104119      0
#> 598   0.067736671      0
#> 599   0.084326875      0
#> 600   1.048559317      0
#> 601   0.855566089      0
#> 602   1.153891572      0
#> 603   1.857664831      0
#> 604   1.228940486      0
#> 605   0.516634041      0
#> 606  -0.350273961      0
#> 607  -0.049549906      0
#> 608  -0.500682330      0
#> 609  -0.237135836      0
#> 610   0.486765465      0
#> 611   2.276097410      0
#> 612   0.042689780      0
#> 613  -0.941699253      0
#> 614   1.052049628      0
#> 615  -1.064600456      0
#> 616   1.371526090      0
#> 617  -0.512467122      0
#> 618  -1.745586731      0
#> 619   0.644886444      0
#> 620  -0.038487624      0
#> 621  -1.659904503      0
#> 622   1.246469415      0
#> 623   0.134161410      0
#> 624   1.906831825      0
#> 625   0.375250871      0
#> 626  -1.132928977      0
#> 627  -2.478901890      0
#> 628   0.311389303      0
#> 629  -0.376481413      0
#> 630  -0.140896285      0
#> 631   1.650360534      0
#> 632   0.157461488      0
#> 633   0.260013263      0
#> 634  -0.406144678      0
#> 635  -0.882677457      0
#> 636  -2.532033474      0
#> 637  -0.145212051      0
#> 638  -1.082576514      0
#> 639  -0.678305537      0
#> 640  -0.041807018      0
#> 641   1.484002991      0
#> 642   0.567702269      0
#> 643  -0.252791296      0
#> 644  -0.271696213      0
#> 645  -0.007899970      0
#> 646  -0.660000914      0
#> 647  -0.911273318      0
#> 648  -1.142399719      0
#> 649   0.865977878      0
#> 650  -1.395063817      0
#> 651   0.768053095      0
#> 652   1.234801265      0
#> 653  -0.169046266      0
#> 654  -0.920994542      0
#> 655   1.204509961      0
#> 656  -0.777993405      0
#> 657  -0.013921152      0
#> 658  -2.133587939      0
#> 659  -0.268633905      0
#> 660  -0.597619932      0
#> 661  -0.351838065      0
#> 662   0.247017754      0
#> 663  -0.597016444      0
#> 664  -1.283853161      0
#> 665  -0.144410016      0
#> 666   0.862733184      0
#> 667   0.071585124      0
#> 668  -0.088810619      0
#> 669   1.502257583      0
#> 670  -0.844170991      0
#> 671   1.257539515      0
#> 672  -1.611885236      0
#> 673  -0.093242734      0
#> 674   0.339921218      0
#> 675   0.375163351      0
#> 676  -0.299773819      0
#> 677  -2.071791200      0
#> 678  -0.505445163      0
#> 679   0.070733432      0
#> 680   0.005286246      0
#> 681  -1.181195963      0
#> 682   0.352707160      0
#> 683   0.279771589      0
#> 684   0.382700993      0
#> 685  -0.523765076      0
#> 686  -1.069418162      0
#> 687  -1.135662738      0
#> 688  -0.811949110      0
#> 689   2.404793385      0
#> 690   0.892310646      0
#> 691   0.112822373      0
#> 692   0.743605244      0
#> 693  -0.948357680      0
#> 694   0.756828437      0
#> 695   0.368499847      0
#> 696  -0.772702643      0
#> 697  -0.684589509      0
#> 698  -0.724337961      0
#> 699  -0.991616380      0
#> 700   0.480490895      0
#> 701  -0.730726247      0
#> 702  -0.839407310      0
#> 703   0.275533570      0
#> 704  -1.872108835      0
#> 705  -0.457516712      0
#> 706  -0.603149513      0
#> 707  -0.689527724      0
#> 708  -1.896271990      0
#> 709   0.171949130      0
#> 710   1.217397318      0
#> 711  -0.079828055      0
#> 712   0.618512543      0
#> 713   0.924960509      0
#> 714   0.300259281      0
#> 715   1.136844043      0
#> 716  -1.478999929      0
#> 717  -0.053310822      0
#> 718  -0.044961706      0
#> 719  -0.440072121      0
#> 720  -1.407486679      0
#> 721  -1.944089097      0
#> 722   0.059595961      0
#> 723   1.144420490      0
#> 724   0.192532013      0
#> 725   0.176204600      0
#> 726   0.008409577      0
#> 727   0.810573410      0
#> 728   0.291557955      0
#> 729  -1.797627601      0
#> 730  -0.256558404      0
#> 731   0.805434019      0
#> 732  -0.705480318      0
#> 733   1.254046947      0
#> 734   0.077345069      0
#> 735  -0.481979044      0
#> 736   0.529650147      0
#> 737   0.357711410      0
#> 738  -0.397550712      0
#> 739   0.566548479      0
#> 740   0.833676222      0
#> 741   0.157747735      0
#> 742   0.801517089      0
#> 743   0.446639404      0
#> 744  -1.972606490      0
#> 745   0.070034128      0
#> 746  -0.175672919      0
#> 747   0.488328032      0
#> 748  -0.666230095      0
#> 749   0.225277744      0
#> 750   0.487703033      0
#> 751   1.457476363      0
#> 752   0.250908851      0
#> 753   0.232847504      0
#> 754  -1.052151269      0
#> 755  -0.339272881      0
#> 756  -1.677984391      0
#> 757   0.514283149      0
#> 758  -0.566656708      0
#> 759   0.630950886      0
#> 760  -0.377033998      0
#> 761   0.221172991      0
#> 762   1.204534512      0
#> 763  -1.971161699      0
#> 764   2.389786964      0
#> 765  -0.287364251      0
#> 766   0.800729501      0
#> 767   1.052277441      0
#> 768   0.630958785      0
#> 769  -0.538861328      0
#> 770  -1.389520780      0
#> 771   0.913064282      0
#> 772  -0.148291564      0
#> 773  -0.601620281      0
#> 774  -0.111365590      0
#> 775   1.014238713      0
#> 776  -0.469581498      0
#> 777   2.272000711      0
#> 778  -1.694297937      0
#> 779   0.322059028      0
#> 780   0.747274542      0
#> 781  -0.303058361      0
#> 782  -0.886144621      0
#> 783   1.575071983      0
#> 784   0.361333706      0
#> 785   0.145678858      0
#> 786  -0.209395371      0
#> 787   0.066283373      0
#> 788   0.574641383      0
#> 789   1.391438830      0
#> 790  -1.322867377      0
#> 791  -1.651153695      0
#> 792   0.920277932      0
#> 793  -0.984902830      0
#> 794  -0.425357679      0
#> 795  -0.750245573      0
#> 796  -0.091449356      0
#> 797   0.206996018      0
#> 798   0.743901999      0
#> 799   0.628454945      0
#> 800   0.248520633      0
#> 801   2.170850499      0
#> 802  -0.862668023      0
#> 803  -0.031927267      0
#> 804   0.410858843      0
#> 805  -0.016601568      0
#> 806   0.160677828      0
#> 807   0.486092075      0
#> 808   1.370654162      0
#> 809  -1.592293572      0
#> 810   0.138371911      0
#> 811   1.141829803      0
#> 812  -0.084625902      0
#> 813   0.363960342      0
#> 814  -1.438159112      0
#> 815   1.266785566      0
#> 816   0.438449100      0
#> 817  -0.201111002      0
#> 818  -0.199719452      0
#> 819   0.065143249      0
#> 820  -0.631755027      0
#> 821  -0.284440124      0
#> 822  -2.086354970      0
#> 823   1.229757753      0
#> 824   0.355933866      0
#> 825   0.538198648      0
#> 826  -0.138836258      0
#> 827   0.459083421      0
#> 828  -0.343377918      0
#> 829   0.521790106      0
#> 830   0.105067503      0
#> 831   1.742797364      0
#> 832  -0.135620259      0
#> 833  -0.686348628      0
#> 834  -0.082738245      0
#> 835   0.714381222      0
#> 836   0.014384058      0
#> 837  -1.003213379      0
#> 838  -1.505825604      0
#> 839  -1.532681052      0
#> 840   1.168809797      0
#> 841  -0.365753470      0
#> 842  -0.539482205      0
#> 843   1.557107000      0
#> 844   1.258776256      0
#> 845  -0.955774929      0
#> 846   0.695457814      0
#> 847   0.609194634      0
#> 848  -0.336818155      0
#> 849  -0.252256365      0
#> 850   0.084599713      0
#> 851   0.640568170      0
#> 852  -0.136222745      0
#> 853   0.392091632      0
#> 854   0.090489993      0
#> 855   0.408160086      0
#> 856   1.483825373      0
#> 857   0.505925668      0
#> 858  -0.524130062      0
#> 859  -0.046868512      0
#> 860  -1.543685504      0
#> 861   0.202184869      0
#> 862  -0.437038830      0
#> 863   0.245561688      0
#> 864   0.028311739      0
#> 865   0.580191702      0
#> 866  -0.745189364      0
#> 867  -0.595386930      0
#> 868  -0.728609648      0
#> 869  -0.132266975      0
#> 870  -0.741687090      0
#> 871   1.051548001      0
#> 872   0.609347247      0
#> 873  -0.639692295      0
#> 874  -0.660023231      0
#> 875   0.111059607      0
#> 876   0.254629109      0
#> 877  -0.913736052      0
#> 878   1.175639297      0
#> 879  -0.425394372      0
#> 880  -3.232104353      0
#> 881   0.180340678      0
#> 882  -0.643647019      0
#> 883  -0.857555994      0
#> 884  -0.222968082      0
#> 885  -0.527874852      0
#> 886  -1.915864393      0
#> 887   1.153158382      0
#> 888  -0.972478543      0
#> 889  -0.424667876      1
#> 890   1.044674349      1
#> 891   0.164693974      1
#> 892  -0.535142766      1
#> 893  -0.907731296      1
#> 894   2.383408136      1
#> 895   0.674649249      1
#> 896  -0.204458811      1
#> 897   0.627565729      1
#> 898  -0.740801186      1
#> 899   2.662476381      1
#> 900   0.176511625      1
#> 901   1.304774477      1
#> 902   0.108547262      1
#> 903  -0.154068384      1
#> 904   3.150749185      1
#> 905  -0.556350766      1
#> 906   1.328844012      1
#> 907   1.253930996      1
#> 908   1.600640596      1
#> 909  -1.123047351      1
#> 910   0.054676632      1
#> 911   0.159921512      1
#> 912  -1.652413594      1
#> 913  -0.455571031      1
#> 914   1.016605842      1
#> 915   1.119587317      1
#> 916  -0.360542974      1
#> 917   0.825067603      1
#> 918  -0.529615740      1
#> 919   2.280154799      1
#> 920   1.248609661      1
#> 921  -0.969564923      1
#> 922   1.450080706      1
#> 923   0.882577477      1
#> 924  -0.673886959      1
#> 925   0.795337561      1
#> 926   1.115735915      1
#> 927   0.536197218      1
#> 928   1.238928563      1
#> 929   0.022818443      1
#> 930  -0.046694217      1
#> 931   1.002451153      1
#> 932   1.474923572      1
#> 933  -1.108646926      1
#> 934   0.587935380      1
#> 935   0.066152097      1
#> 936  -1.182515623      1
#> 937  -0.062374175      1
#> 938   2.137818507      1
#> 939   1.753381429      1
#> 940   2.123517257      1
#> 941   1.957295200      1
#> 942   0.574874989      1
#> 943   1.614030577      1
#> 944   0.712804271      1
#> 945  -0.343403603      1
#> 946   0.203157910      1
#> 947   0.733491802      1
#> 948   0.417435774      1
#> 949   0.897438151      1
#> 950   1.974273875      1
#> 951   0.510506481      1
#> 952   0.611035719      1
#> 953   0.388983306      1
#> 954   0.074698839      1
#> 955   0.685350501      1
#> 956  -0.511153296      1
#> 957   1.190432558      1
#> 958   0.023150393      1
#> 959   2.972345287      1
#> 960   1.457002921      1
#> 961  -0.627768854      1
#> 962   1.083303454      1
#> 963   1.842928315      1
#> 964   1.786462218      1
#> 965   0.568121043      1
#> 966   1.206895348      1
#> 967   0.612339363      1
#> 968  -1.045923497      1
#> 969  -1.280265154      1
#> 970  -0.129353845      1
#> 971  -0.595401820      1
#> 972  -0.280614127      1
#> 973  -0.292668943      1
#> 974   1.414671911      1
#> 975   1.756279707      1
#> 976   0.619274159      1
#> 977   0.294727235      1
#> 978   0.244939435      1
#> 979   1.675932790      1
#> 980   0.684411632      1
#> 981   1.202989594      1
#> 982   2.246363526      1
#> 983  -0.440710261      1
#> 984   0.107217066      1
#> 985  -0.617976328      1
#> 986  -0.380190439      1
#> 987   2.310328443      1
#> 988   0.099825286      1
#> 989   0.627036360      1
#> 990   1.638867828      1
#> 991   1.892669042      1
#> 992  -0.967379755      1
#> 993  -0.621645361      1
#> 994  -0.607467740      1
#> 995  -0.255202863      1
#> 996   1.999848035      1
#> 997   1.006445147      1
#> 998   0.951567262      1
#> 999   1.017053678      1
#> 1000 -0.424457244      1