Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -0.4894393321      0
#> 2     1.4575295924      0
#> 3    -1.3623520375      0
#> 4     0.0385202846      0
#> 5     0.9044707536      0
#> 6     0.5517286065      0
#> 7    -0.6799827822      0
#> 8    -1.2212706772      0
#> 9    -2.0854843833      0
#> 10   -0.9085601763      0
#> 11    0.3912347456      0
#> 12   -0.6425083259      0
#> 13    0.6629711864      0
#> 14   -1.0970882714      0
#> 15    0.4164033224      0
#> 16   -1.0233990266      0
#> 17   -1.3882603780      0
#> 18    1.2770695097      0
#> 19   -0.2581986676      0
#> 20   -0.5505141664      0
#> 21   -0.0718094254      0
#> 22   -0.8104357356      0
#> 23   -0.8540212357      0
#> 24    0.9979511408      0
#> 25    1.3212146397      0
#> 26    0.3474225944      0
#> 27    0.1786121978      0
#> 28   -0.2517340341      0
#> 29   -0.4062703014      0
#> 30    0.4692927965      0
#> 31    0.9543926902      0
#> 32   -0.1250414063      0
#> 33    1.0734363332      0
#> 34    0.2572223369      0
#> 35    0.1906286104      0
#> 36   -0.1230262335      0
#> 37    1.0596963524      0
#> 38   -0.6590534506      0
#> 39    0.0673954711      0
#> 40    1.0575776431      0
#> 41    1.4496958171      0
#> 42   -0.4992792038      0
#> 43    2.1829562607      0
#> 44   -0.6253437083      0
#> 45   -1.6186538167      0
#> 46   -0.0959177534      0
#> 47    0.4193262624      0
#> 48    0.2890971425      0
#> 49    1.3136003109      0
#> 50   -0.5071447026      0
#> 51   -1.3556779013      0
#> 52    0.5172243543      0
#> 53    0.1422442956      0
#> 54   -0.1704500001      0
#> 55   -0.7100350949      0
#> 56    0.7150733844      0
#> 57   -2.5966778172      0
#> 58   -0.7938954128      0
#> 59    0.7215685253      0
#> 60   -0.4450781642      0
#> 61    1.0929978085      0
#> 62    0.4814159810      0
#> 63   -0.4568781935      0
#> 64   -0.6187439475      0
#> 65   -0.1795882142      0
#> 66    0.2157409063      0
#> 67   -0.6854444096      0
#> 68   -0.0653032371      0
#> 69    2.0323438433      0
#> 70   -0.3910264473      0
#> 71    0.3534127257      0
#> 72    0.5371463980      0
#> 73   -0.0183777665      0
#> 74   -0.4382908932      0
#> 75   -1.7008656594      0
#> 76   -0.7838398861      0
#> 77    1.6689289784      0
#> 78   -0.1214366377      0
#> 79   -0.2618632956      0
#> 80   -0.2382058728      0
#> 81    1.7611372751      0
#> 82    0.4801382627      0
#> 83   -1.5452894815      0
#> 84    0.5432994895      0
#> 85   -0.4699081307      0
#> 86    0.6825008515      0
#> 87   -0.7630336694      0
#> 88   -1.3617809471      0
#> 89   -0.0045761009      0
#> 90    1.3763126025      0
#> 91   -0.4422758734      0
#> 92   -1.0914418351      0
#> 93   -0.6853927871      0
#> 94   -0.8664863543      0
#> 95   -1.1039788416      0
#> 96   -0.6443465844      0
#> 97   -0.0949575138      0
#> 98   -0.4351740111      0
#> 99    0.7919471383      0
#> 100  -0.3773156929      0
#> 101  -0.9585114697      0
#> 102  -0.0014371680      0
#> 103   0.2851711389      0
#> 104   0.5360666421      0
#> 105  -1.0095533667      0
#> 106  -0.0746535608      0
#> 107  -1.5953845282      0
#> 108  -1.0277515723      0
#> 109  -0.2506311945      0
#> 110   1.4181505028      0
#> 111  -0.7284310591      0
#> 112   0.8516301085      0
#> 113   0.3457768633      0
#> 114   0.1661904268      0
#> 115   0.6310203362      0
#> 116  -0.2518147247      0
#> 117  -0.9246371636      0
#> 118  -0.8144120235      0
#> 119  -0.6084261574      0
#> 120  -0.5187942135      0
#> 121   0.6481358732      0
#> 122  -0.4948753776      0
#> 123   1.5699823923      0
#> 124   0.3798824820      0
#> 125  -0.1588081001      0
#> 126  -1.6380692797      0
#> 127  -1.2384780859      0
#> 128  -0.1662239005      0
#> 129   1.0163672753      0
#> 130   0.5640149060      0
#> 131   0.1764904536      0
#> 132  -0.1586005539      0
#> 133   0.3862899504      0
#> 134   0.9415049920      0
#> 135   0.6371336504      0
#> 136   0.2087972511      0
#> 137   0.4629701572      0
#> 138  -1.6611015063      0
#> 139  -0.0707442677      0
#> 140   0.8801147280      0
#> 141  -0.5610847622      0
#> 142  -0.2688204158      0
#> 143  -0.6043773958      0
#> 144  -0.4343380621      0
#> 145   0.4521423854      0
#> 146   1.1166186687      0
#> 147  -1.1718159624      0
#> 148   1.1644724294      0
#> 149   0.2440101555      0
#> 150  -0.2575265241      0
#> 151   0.3845256688      0
#> 152  -1.0380757938      0
#> 153  -0.8174431059      0
#> 154   0.9643842906      0
#> 155   0.4468106640      0
#> 156  -0.2705642883      0
#> 157   0.0280026162      0
#> 158  -0.9792951744      0
#> 159   1.4909390894      0
#> 160  -0.4664699222      0
#> 161  -1.0453989525      0
#> 162   1.1911618327      0
#> 163  -0.7030557207      0
#> 164   1.5200450692      0
#> 165   1.1793985757      0
#> 166  -1.4607471446      0
#> 167  -0.0260226273      0
#> 168   1.1493718067      0
#> 169   1.0062070533      0
#> 170  -0.5523141893      0
#> 171  -0.0615335202      0
#> 172   0.3255634839      0
#> 173  -1.3230099817      0
#> 174   0.2450531403      0
#> 175  -0.8364349479      0
#> 176   1.4572672377      0
#> 177   2.1184267588      0
#> 178   0.1615593864      0
#> 179   0.5441996526      0
#> 180  -0.0229335799      0
#> 181  -0.8342064664      0
#> 182  -0.9786451266      0
#> 183   1.4974217168      0
#> 184  -0.7394768166      0
#> 185  -0.6684759468      0
#> 186  -1.4459046227      0
#> 187  -0.8066524806      0
#> 188  -0.9075598846      0
#> 189   1.2492740043      0
#> 190  -0.0849480517      0
#> 191   1.8280130628      0
#> 192  -0.2983713684      0
#> 193  -1.0303499901      0
#> 194   1.3837010964      0
#> 195  -0.0377774691      0
#> 196  -0.1844730105      0
#> 197   1.5024982845      0
#> 198  -0.3824391318      0
#> 199  -0.2703613037      0
#> 200  -0.0883955515      0
#> 201  -0.6755845483      0
#> 202   0.7186301427      0
#> 203  -0.4746275110      0
#> 204   1.4400548797      0
#> 205   0.1346011685      0
#> 206   0.1844747186      0
#> 207   0.0351379376      0
#> 208   1.3954737049      0
#> 209  -0.5077472718      0
#> 210  -0.2227665475      0
#> 211  -0.5694405374      0
#> 212   1.1202681662      0
#> 213  -0.9293947349      0
#> 214  -0.6722222162      0
#> 215  -0.4215553938      0
#> 216   0.1474500083      0
#> 217  -1.0032251176      0
#> 218   0.5714057930      0
#> 219  -0.1155260157      0
#> 220  -0.0415704813      0
#> 221   0.3401077967      0
#> 222  -0.7982891903      0
#> 223   0.3621079104      0
#> 224   2.1378793718      0
#> 225  -2.3140337270      0
#> 226  -0.8742505614      0
#> 227   1.0921005975      0
#> 228  -0.5982347257      0
#> 229   0.5843219725      0
#> 230  -0.3394706840      0
#> 231   0.3258135763      0
#> 232   0.9760745135      0
#> 233  -1.7016515448      0
#> 234  -1.1735333286      0
#> 235   0.6354183533      0
#> 236  -0.9479252831      0
#> 237   1.6219418610      0
#> 238  -0.7401354803      0
#> 239   1.0407921724      0
#> 240  -0.1871640686      0
#> 241  -0.2886023184      0
#> 242  -0.6463115153      0
#> 243  -0.1012870735      0
#> 244  -0.4578273232      0
#> 245  -0.8068390394      0
#> 246  -0.3295852880      0
#> 247   0.2295168434      0
#> 248   0.2320411836      0
#> 249   1.0875781673      0
#> 250   1.8366276071      0
#> 251  -0.0155551052      0
#> 252  -0.8332083970      0
#> 253  -0.3252902560      0
#> 254  -1.4634475708      0
#> 255   0.3590355245      0
#> 256  -0.4631894944      0
#> 257  -0.0234226095      0
#> 258   1.2121260243      0
#> 259   0.6135243599      0
#> 260  -0.9128321284      0
#> 261  -1.0636773249      0
#> 262  -0.1045017189      0
#> 263   1.2268083374      0
#> 264  -0.2638378964      0
#> 265  -1.0920811806      0
#> 266  -1.7618888050      0
#> 267  -0.5777842484      0
#> 268   1.0768439809      0
#> 269  -0.5113228856      0
#> 270   0.4317005766      0
#> 271  -1.6943928247      0
#> 272  -0.2800907667      0
#> 273   0.4737690831      0
#> 274   1.4248280993      0
#> 275   0.9787370034      0
#> 276   0.7469392784      0
#> 277  -0.5931900818      0
#> 278  -1.1952004545      0
#> 279   0.0109245559      0
#> 280  -0.7453870641      0
#> 281  -0.0692911829      0
#> 282   0.2498683108      0
#> 283  -0.3702069046      0
#> 284   0.9183988953      0
#> 285  -0.1767124576      0
#> 286   0.5554508797      0
#> 287  -0.5176465318      0
#> 288   0.1752283521      0
#> 289   1.5292763933      0
#> 290   2.0423708992      0
#> 291   1.6441059435      0
#> 292   2.6985392177      0
#> 293  -0.1268046705      0
#> 294  -0.1610049467      0
#> 295   0.1785473083      0
#> 296   0.1202472054      0
#> 297  -0.9004331281      0
#> 298   0.4101731005      0
#> 299   1.9294077459      0
#> 300   0.8601790409      0
#> 301  -0.2416015284      0
#> 302   0.1559678881      0
#> 303  -0.4880096678      0
#> 304  -0.2534545924      0
#> 305  -0.9181571781      0
#> 306  -0.4309506757      0
#> 307  -0.3157849053      0
#> 308   0.1622307636      0
#> 309   1.5865473731      0
#> 310   0.2730433230      0
#> 311  -1.1760712485      0
#> 312  -0.9626664484      0
#> 313   1.7019814234      0
#> 314  -0.8607727094      0
#> 315   0.1430035155      0
#> 316   1.6762641814      0
#> 317  -0.9690688102      0
#> 318   0.8268666011      0
#> 319   0.6542225624      0
#> 320   1.1696720232      0
#> 321  -1.4860275455      0
#> 322  -0.8635632050      0
#> 323  -1.1610560474      0
#> 324  -1.0555921347      0
#> 325   2.1045480081      0
#> 326   3.0961339375      0
#> 327   0.0838134804      0
#> 328  -0.0207480512      0
#> 329  -0.5152122382      0
#> 330  -0.5230671604      0
#> 331  -1.1639547318      0
#> 332  -0.5611980487      0
#> 333   1.5097979027      0
#> 334  -0.7303900747      0
#> 335  -0.9004787678      0
#> 336  -0.1700102209      0
#> 337   1.0364642490      0
#> 338   1.7370943751      0
#> 339  -0.9348734491      0
#> 340  -2.1063253926      0
#> 341   0.6837916009      0
#> 342   0.4371394906      0
#> 343  -0.2893048968      0
#> 344  -0.5719173872      0
#> 345  -0.9851857560      0
#> 346  -0.3873400816      0
#> 347  -1.3259258486      0
#> 348  -1.1855351510      0
#> 349  -0.1590700654      0
#> 350   1.5933795923      0
#> 351   1.0821866322      0
#> 352   1.6668252339      0
#> 353  -0.0004030580      0
#> 354   0.8265001850      0
#> 355   0.5144892349      0
#> 356  -0.1522848904      0
#> 357  -0.9419155086      0
#> 358   0.0046232267      0
#> 359  -0.7319918775      0
#> 360  -1.1371384571      0
#> 361  -1.2232564959      0
#> 362   0.1200471403      0
#> 363  -0.3690736805      0
#> 364  -0.6292949225      0
#> 365   0.6427170233      0
#> 366  -0.5008121932      0
#> 367   0.8756859837      0
#> 368  -0.7558857454      0
#> 369   0.1383837879      0
#> 370   1.1950150382      0
#> 371  -1.4354058336      0
#> 372   2.6175836220      0
#> 373  -0.2361028954      0
#> 374  -0.3607173186      0
#> 375  -0.6414027490      0
#> 376   0.1874371998      0
#> 377   0.0932923415      0
#> 378  -1.4486323871      0
#> 379  -0.3012387186      0
#> 380  -0.8619017129      0
#> 381   0.3789710096      0
#> 382  -0.0389077195      0
#> 383  -1.0038941984      0
#> 384  -1.5480476212      0
#> 385  -0.4392513456      0
#> 386   0.8844288333      0
#> 387   0.0173499801      0
#> 388   0.2899363679      0
#> 389  -0.0487958721      0
#> 390  -0.6964808614      0
#> 391  -0.2835305057      0
#> 392  -0.4556204694      0
#> 393  -0.1343360852      0
#> 394   0.5969343713      0
#> 395  -0.4585049583      0
#> 396   1.4601285618      0
#> 397  -0.6040943718      0
#> 398  -1.0875642676      0
#> 399  -1.2756376180      0
#> 400   1.8302968254      0
#> 401  -0.3703164886      0
#> 402   1.0797142616      0
#> 403   1.6782540747      0
#> 404  -0.2893035486      0
#> 405  -1.1925316227      0
#> 406  -0.2992654683      0
#> 407  -1.3326763954      0
#> 408   0.2026649659      0
#> 409   0.1582375924      0
#> 410   0.4956327259      0
#> 411   0.0961862623      0
#> 412   1.1573001399      0
#> 413   0.1444112458      0
#> 414  -1.0899385509      0
#> 415   0.3550656307      0
#> 416  -1.5403742875      0
#> 417   0.3936978023      0
#> 418  -0.0966628518      0
#> 419  -0.8110510843      0
#> 420   1.1569075775      0
#> 421   0.3799334433      0
#> 422   1.0305206956      0
#> 423   0.8382602911      0
#> 424  -0.8090164597      0
#> 425   2.0877668147      0
#> 426   0.3846169642      0
#> 427  -1.4008327956      0
#> 428   0.6422252599      0
#> 429   0.2099206122      0
#> 430  -0.4928690477      0
#> 431  -0.3122552270      0
#> 432  -0.9245039566      0
#> 433  -0.3291235653      0
#> 434  -1.0063646594      0
#> 435  -0.1911740735      0
#> 436  -0.1132127782      0
#> 437  -0.3547991495      0
#> 438  -0.0560453182      0
#> 439   1.7534875568      0
#> 440  -0.7213559385      0
#> 441  -0.8739359696      0
#> 442   2.7361008433      0
#> 443   0.5429484146      0
#> 444  -0.7731578265      0
#> 445  -0.5522475930      0
#> 446   1.2809250969      0
#> 447   1.2899148937      0
#> 448  -1.6146635942      0
#> 449  -0.7272137365      0
#> 450   1.4088485660      0
#> 451  -2.5624851657      0
#> 452   0.2689238954      0
#> 453  -1.5091293707      0
#> 454   0.8866153904      0
#> 455  -0.0478794704      0
#> 456   0.6996129796      0
#> 457  -0.9756559136      0
#> 458   0.6154581912      0
#> 459   0.5466413494      0
#> 460   0.0134020940      0
#> 461  -0.8586558909      0
#> 462  -1.0611135407      0
#> 463  -0.7136682430      0
#> 464  -0.1345652982      0
#> 465  -0.1494820290      0
#> 466   0.0608970812      0
#> 467  -0.4370353561      0
#> 468  -1.5510848839      0
#> 469   0.1908435579      0
#> 470   0.4760157109      0
#> 471   1.4614568404      0
#> 472  -1.1356056564      0
#> 473   0.0236558371      0
#> 474  -2.3884698757      0
#> 475   0.8464901829      0
#> 476  -0.5943897404      0
#> 477  -0.3724149911      0
#> 478  -0.5601811412      0
#> 479  -1.8703785832      0
#> 480   0.7010717447      0
#> 481  -0.0500537741      0
#> 482  -0.2528979278      0
#> 483   1.1877619735      0
#> 484   1.9409847307      0
#> 485   1.5753212886      0
#> 486   0.1611741511      0
#> 487   0.2684297339      0
#> 488  -0.3482304577      0
#> 489  -0.3082426367      0
#> 490  -0.8684519032      0
#> 491   0.1320173179      0
#> 492  -3.4997942138      0
#> 493   0.7837668397      0
#> 494  -1.2445562727      0
#> 495  -0.2923624037      0
#> 496   1.3509799683      0
#> 497  -0.8802794800      0
#> 498   0.0109306473      0
#> 499   1.1279454575      0
#> 500  -0.1439270072      0
#> 501  -0.2603039817      0
#> 502  -2.1707908090      0
#> 503  -2.5697121277      0
#> 504   0.9410094973      0
#> 505   0.5004827597      0
#> 506  -0.2782071444      0
#> 507   0.3960126523      0
#> 508   0.9054702304      0
#> 509  -0.6507552825      0
#> 510   0.1640616887      0
#> 511  -0.6602586113      0
#> 512  -2.0370510746      0
#> 513  -0.0614215840      0
#> 514   1.7973653871      0
#> 515   0.1160864225      0
#> 516   1.5374183057      0
#> 517  -0.5750244999      0
#> 518   0.7386025857      0
#> 519  -1.1383827740      0
#> 520  -0.1816943616      0
#> 521   0.1464001503      0
#> 522   0.3403454362      0
#> 523   0.2123644377      0
#> 524   0.6111699412      0
#> 525   1.1777689104      0
#> 526  -0.5748424834      0
#> 527  -1.2890489339      0
#> 528   1.3033891415      0
#> 529   0.7199690226      0
#> 530   0.4038615567      0
#> 531   0.4187938540      0
#> 532   1.9038430303      0
#> 533   0.9814925083      0
#> 534   0.9315177805      0
#> 535  -0.3374433445      0
#> 536  -0.6544167272      0
#> 537   0.5344427497      0
#> 538  -0.3072229126      0
#> 539  -0.2241008957      0
#> 540  -0.7943685271      0
#> 541   0.3657786615      0
#> 542  -0.3146014418      0
#> 543   0.0973488227      0
#> 544   1.7517755693      0
#> 545  -0.9816901383      0
#> 546  -3.1056970911      0
#> 547   1.0137601961      0
#> 548   0.1162164740      0
#> 549  -0.2442988375      0
#> 550  -0.2590076397      0
#> 551  -0.4006994731      0
#> 552   0.0228264596      0
#> 553   1.2632581789      0
#> 554   0.4163161215      0
#> 555  -2.0609460077      0
#> 556  -0.3335221700      0
#> 557   0.5848352033      0
#> 558  -0.3246941566      0
#> 559   2.7173024546      0
#> 560   1.6014901492      0
#> 561   2.8223311210      0
#> 562   1.1808903769      0
#> 563  -0.7156753948      0
#> 564   1.0636629217      0
#> 565  -2.3220486495      0
#> 566  -0.5595932801      0
#> 567  -1.8498417434      0
#> 568  -0.0012744426      0
#> 569  -0.8071205234      0
#> 570  -0.0480277865      0
#> 571   0.2441793216      0
#> 572   0.3175197065      0
#> 573   0.3028676032      0
#> 574  -0.4421016904      0
#> 575  -0.3078548278      0
#> 576   0.4468523355      0
#> 577  -1.7417671719      0
#> 578  -1.1995531094      0
#> 579  -0.5941767345      0
#> 580  -2.0188177714      0
#> 581   0.9802875097      0
#> 582   0.6846622455      0
#> 583   0.6228855422      0
#> 584   1.2401792119      0
#> 585   0.2426814250      0
#> 586   0.0701029744      0
#> 587   0.3430315298      0
#> 588   1.0271392520      0
#> 589   1.0474103099      0
#> 590  -0.5779203153      0
#> 591  -1.5378166649      0
#> 592   0.3123571854      0
#> 593  -0.9070826879      0
#> 594  -0.3444498466      0
#> 595   0.0146989652      0
#> 596  -0.0168760805      0
#> 597   1.3878767794      0
#> 598  -1.9185754905      0
#> 599   0.2344066697      0
#> 600   0.9256767039      0
#> 601   1.7899101606      0
#> 602   0.2138975625      0
#> 603  -0.3357126918      0
#> 604   1.1719061717      0
#> 605  -0.6685847556      0
#> 606   0.3588038671      0
#> 607  -0.5675171946      0
#> 608  -1.7181021103      0
#> 609  -1.5657706990      0
#> 610   0.4153816041      0
#> 611   0.1288456280      0
#> 612  -0.2583805806      0
#> 613  -0.3763494522      0
#> 614  -0.7623341284      0
#> 615   0.8579532816      0
#> 616  -1.2699989530      0
#> 617   0.7816380360      0
#> 618  -0.3221343642      0
#> 619   2.1665706616      0
#> 620  -0.6764732271      0
#> 621   0.2253920152      0
#> 622  -1.1981818711      0
#> 623   0.9438216110      0
#> 624  -0.4937591207      0
#> 625   0.2551340812      0
#> 626  -0.2746491450      0
#> 627   0.8772213829      0
#> 628  -1.7955022879      0
#> 629   1.3537589346      0
#> 630  -0.9596090599      0
#> 631  -0.6880475171      0
#> 632   0.8201365906      0
#> 633   0.5498079373      0
#> 634  -1.1985788786      0
#> 635   0.8284022866      0
#> 636  -1.0386528847      0
#> 637   0.3934927700      0
#> 638   0.6066132793      0
#> 639   2.5803184522      0
#> 640  -0.7040990641      0
#> 641  -1.3614446272      0
#> 642  -1.6466978566      0
#> 643  -0.3021176452      0
#> 644   1.7982847923      0
#> 645   1.7968032994      0
#> 646   0.7095826437      0
#> 647   0.1541278246      0
#> 648  -0.1504953949      0
#> 649   0.9106103079      0
#> 650   1.0580407119      0
#> 651  -0.0002086229      0
#> 652   1.5380769643      0
#> 653   0.1600205667      0
#> 654   1.1952827663      0
#> 655   2.6725709067      0
#> 656  -0.9510921475      0
#> 657  -1.1397364722      0
#> 658   0.0460987481      0
#> 659  -0.1771391486      0
#> 660  -0.1097905562      0
#> 661  -1.3236124996      0
#> 662   0.6360213951      0
#> 663   0.2920761594      0
#> 664   1.6615102248      0
#> 665   0.1418099931      0
#> 666   0.6164766509      0
#> 667   0.9359634271      0
#> 668   1.3808903914      0
#> 669   0.3490970316      0
#> 670  -0.2991352181      0
#> 671  -0.2570976682      0
#> 672   0.2998853766      0
#> 673   0.5883211846      0
#> 674  -0.8368508043      0
#> 675   1.1735339409      0
#> 676  -0.4875474616      0
#> 677  -0.6375671598      0
#> 678  -0.4858910416      0
#> 679  -0.3112426108      0
#> 680   1.4590666440      0
#> 681  -0.5415739277      0
#> 682   1.5090695168      0
#> 683   0.9316747038      0
#> 684   1.2938978537      0
#> 685   0.8573359494      0
#> 686   0.4082773830      0
#> 687  -0.3989852072      0
#> 688   0.3168379126      0
#> 689   0.3586901000      0
#> 690  -0.0995681427      0
#> 691   1.1418202599      0
#> 692  -0.5033465783      0
#> 693  -0.1605519210      0
#> 694  -0.4757939251      0
#> 695  -0.0944263086      0
#> 696  -0.3227184678      0
#> 697  -1.7575879428      0
#> 698  -1.0773646734      0
#> 699  -0.4210399152      0
#> 700  -0.7400079272      0
#> 701  -0.6234006674      0
#> 702   0.2510584188      0
#> 703   0.1775526888      0
#> 704   0.4316442031      0
#> 705  -0.6139190545      0
#> 706   0.3061090892      0
#> 707   0.5731693541      0
#> 708  -1.0520962819      0
#> 709  -0.8206401754      0
#> 710   0.0109958712      0
#> 711  -0.3194009416      0
#> 712   0.3008286349      0
#> 713   0.7484125304      0
#> 714   0.5214262071      0
#> 715   0.4511944220      0
#> 716   1.2006122051      0
#> 717   0.7190374576      0
#> 718  -1.2825470008      0
#> 719   0.8834785966      0
#> 720   1.2304624846      0
#> 721  -1.1156146159      0
#> 722  -0.1896462916      0
#> 723  -0.6189282232      0
#> 724   0.2521878350      0
#> 725  -0.8165410248      0
#> 726  -1.0631183016      0
#> 727  -0.1627458306      0
#> 728   1.0904367881      0
#> 729  -0.2260848373      0
#> 730  -0.0481214568      0
#> 731   0.8147855676      0
#> 732   0.2985971302      0
#> 733  -0.0367346330      0
#> 734  -1.0915636072      0
#> 735  -0.7872451862      0
#> 736  -0.2517979731      0
#> 737   0.6828401228      0
#> 738  -0.8088779576      0
#> 739   0.3436563811      0
#> 740   0.7987305568      0
#> 741  -0.3591553298      0
#> 742  -0.1725380841      0
#> 743   0.1057490938      0
#> 744  -2.2277153624      0
#> 745  -0.3623748125      0
#> 746   1.8394472881      0
#> 747  -0.7910803584      0
#> 748   0.4639622112      0
#> 749   1.0916973874      0
#> 750   0.5465713142      0
#> 751  -0.8175053783      0
#> 752  -0.3762970381      0
#> 753   1.1766951174      0
#> 754   2.5326554671      0
#> 755   1.4407845525      0
#> 756  -1.4776750052      0
#> 757  -1.1643774886      0
#> 758   0.4949533073      0
#> 759   0.7826426420      0
#> 760  -0.0641033173      0
#> 761  -1.7705234620      0
#> 762  -0.3422291425      0
#> 763   1.0726809151      0
#> 764   0.5970889677      0
#> 765   0.6667088791      0
#> 766  -0.6060204722      0
#> 767  -1.0080589084      0
#> 768  -0.0511744619      0
#> 769   1.1056104143      0
#> 770  -1.3737919265      0
#> 771  -0.4345853616      0
#> 772   3.0094285570      0
#> 773  -1.3678062089      0
#> 774  -1.4999289650      0
#> 775   0.7063443030      0
#> 776  -1.0209208640      0
#> 777   1.2268868473      0
#> 778   0.0630263439      0
#> 779  -0.5753556206      0
#> 780  -0.2691584686      0
#> 781  -1.4420307200      0
#> 782  -0.4992410658      0
#> 783  -0.6473057962      0
#> 784   0.2745082318      0
#> 785   1.5020894626      0
#> 786  -0.2171151303      0
#> 787  -0.0933529355      0
#> 788   0.1197375739      0
#> 789  -0.9408321778      0
#> 790   0.2576746662      0
#> 791   1.3095250369      0
#> 792   0.5771245492      0
#> 793  -1.0949060585      0
#> 794   0.5362246142      0
#> 795   0.3873314105      0
#> 796   0.1169872073      0
#> 797  -0.6698245999      0
#> 798  -0.0449874707      0
#> 799   0.7123696122      0
#> 800  -2.4242293872      0
#> 801  -0.1612016430      0
#> 802   2.0091099542      0
#> 803  -0.8723535270      0
#> 804  -0.2617403948      0
#> 805  -0.1865116402      0
#> 806  -0.9625328633      0
#> 807   1.0941439812      0
#> 808   0.3760805227      0
#> 809   2.0637251687      0
#> 810   2.1218554011      0
#> 811   0.5652143007      0
#> 812   0.5975118083      0
#> 813   0.6721826569      0
#> 814   0.6164559356      0
#> 815   2.0117049959      0
#> 816  -0.5920635370      0
#> 817   0.2971173140      0
#> 818  -0.1408380245      0
#> 819   0.5226455984      0
#> 820  -0.7316640911      0
#> 821   1.6137923392      0
#> 822  -0.0035925308      0
#> 823   0.0977480827      0
#> 824  -0.2115705225      0
#> 825  -0.0458687244      0
#> 826   0.1405237763      0
#> 827   0.0638965960      0
#> 828   0.0520860339      0
#> 829   1.1941952385      0
#> 830   0.9437084675      0
#> 831   0.6966751192      0
#> 832  -2.3415344718      0
#> 833  -0.7928895547      0
#> 834  -0.9796213393      0
#> 835   0.5491172672      0
#> 836   2.7313488798      0
#> 837   0.2480968188      0
#> 838  -0.0408006312      0
#> 839  -0.1665225209      0
#> 840  -0.5347741202      0
#> 841  -0.3009504359      0
#> 842  -0.4852116189      0
#> 843   0.6130687221      0
#> 844   1.0510041432      0
#> 845  -0.4733820268      0
#> 846  -0.5198935541      0
#> 847  -1.5572078814      0
#> 848   0.5741063740      0
#> 849   1.2651946824      0
#> 850  -0.5816455041      0
#> 851  -0.9035448386      0
#> 852   0.6241065465      0
#> 853  -0.5534016707      0
#> 854  -0.3495459694      0
#> 855   0.5897025892      0
#> 856  -0.5697440552      0
#> 857  -0.4524456420      0
#> 858  -0.5378012535      0
#> 859   0.7017947969      0
#> 860  -0.0613853281      0
#> 861  -0.5783337804      0
#> 862   0.6243536584      0
#> 863  -0.1907162638      0
#> 864  -0.8241489354      0
#> 865  -0.4135854091      0
#> 866   0.9500503484      0
#> 867   0.1486314882      0
#> 868  -0.4933398996      0
#> 869  -0.3823782878      0
#> 870  -0.4898270588      0
#> 871   1.1197641960      0
#> 872  -0.6552505653      0
#> 873  -2.8676378200      0
#> 874  -2.3885934077      0
#> 875   0.8797818600      0
#> 876   0.8754106084      0
#> 877  -1.3826873539      0
#> 878   0.6986320501      0
#> 879  -1.2217745063      0
#> 880  -1.1393774691      0
#> 881  -0.0019023562      0
#> 882  -0.3998626564      0
#> 883   0.7598076960      0
#> 884   0.9587555927      0
#> 885   0.2276297116      0
#> 886  -2.7411827576      0
#> 887   0.3978228151      0
#> 888   0.3660147425      0
#> 889  -0.9622938095      0
#> 890   0.6046204084      0
#> 891   0.1702803361      0
#> 892  -0.9555769579      0
#> 893   1.2007585854      0
#> 894  -0.8320768758      0
#> 895  -0.3968049134      0
#> 896   0.3998864114      0
#> 897   0.3164425678      0
#> 898  -0.3933365613      0
#> 899  -0.4357498561      1
#> 900   3.2677892716      1
#> 901   2.0882182994      1
#> 902   0.2278693664      1
#> 903   2.7107891049      1
#> 904   1.9081354084      1
#> 905   2.1195209106      1
#> 906  -0.2400684605      1
#> 907   1.4951247415      1
#> 908   1.4807534862      1
#> 909   0.4627458829      1
#> 910   0.1627902433      1
#> 911   1.1414106944      1
#> 912   1.5163451738      1
#> 913   0.1413332991      1
#> 914   1.4658416401      1
#> 915   1.6609458596      1
#> 916   0.9928697062      1
#> 917   0.1315305536      1
#> 918  -1.4696288781      1
#> 919   0.2403918867      1
#> 920   2.2433397865      1
#> 921   0.1439546499      1
#> 922   0.7172893050      1
#> 923   3.0870453221      1
#> 924   0.1291875883      1
#> 925   0.1198923959      1
#> 926  -1.6185018921      1
#> 927   0.0117299119      1
#> 928   0.4728027837      1
#> 929   0.1256491233      1
#> 930   0.5241555504      1
#> 931  -0.8582126448      1
#> 932  -0.1751339708      1
#> 933   0.5668196139      1
#> 934   0.5727312245      1
#> 935   1.5027908622      1
#> 936   1.6130917364      1
#> 937   0.3676202355      1
#> 938  -0.0675336368      1
#> 939   0.0354868655      1
#> 940   1.7452914047      1
#> 941   0.5711856145      1
#> 942   1.3427118197      1
#> 943   0.0148829545      1
#> 944  -0.5980880964      1
#> 945  -0.5055946802      1
#> 946   1.7968551036      1
#> 947  -0.3572271985      1
#> 948   0.9573830546      1
#> 949   0.7797891344      1
#> 950   1.5717537919      1
#> 951  -1.3446168943      1
#> 952   1.1175447739      1
#> 953  -0.7341723251      1
#> 954   1.3708821269      1
#> 955   0.0255371583      1
#> 956   2.0246765324      1
#> 957   1.6314106880      1
#> 958  -0.3416916933      1
#> 959   1.2705481916      1
#> 960  -0.0161920055      1
#> 961   2.4186278304      1
#> 962   1.6954109577      1
#> 963   1.5977605356      1
#> 964   1.7119581692      1
#> 965   1.9897454152      1
#> 966  -0.5587423153      1
#> 967   1.1637875134      1
#> 968   1.7109098841      1
#> 969   1.0639063877      1
#> 970   0.3091504547      1
#> 971   2.6648631619      1
#> 972  -0.4914488857      1
#> 973   1.4694387053      1
#> 974   0.7966472350      1
#> 975   1.9474609101      1
#> 976  -0.3146442799      1
#> 977   0.6088388630      1
#> 978  -0.0835637978      1
#> 979  -0.3346605680      1
#> 980   0.3611513470      1
#> 981   1.1400093231      1
#> 982   1.8035368222      1
#> 983   0.9396624471      1
#> 984   1.0561843562      1
#> 985   1.5389510629      1
#> 986  -0.9995814232      1
#> 987   1.2020076715      1
#> 988   1.3905109813      1
#> 989   2.3941507608      1
#> 990   0.9981213104      1
#> 991   1.3401125613      1
#> 992  -0.6786561088      1
#> 993  -0.2889239854      1
#> 994   1.4330978248      1
#> 995  -1.3876658309      1
#> 996   1.6309699271      1
#> 997   1.7704410515      1
#> 998   2.0160669967      1
#> 999   0.5515694979      1
#> 1000  2.5768769203      1