Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -0.7804219063      0
#> 2    -0.1354566049      0
#> 3    -0.5942214437      0
#> 4     0.4880656087      0
#> 5    -1.5724888679      0
#> 6    -1.7420413620      0
#> 7    -0.8295101088      0
#> 8    -0.5204692734      0
#> 9     1.1624267402      0
#> 10    0.6664109667      0
#> 11    0.1379076896      0
#> 12    0.1353960069      0
#> 13   -0.7686419716      0
#> 14   -0.8156400095      0
#> 15   -1.2582759319      0
#> 16   -0.3561309906      0
#> 17    0.6789262149      0
#> 18   -0.4706589445      0
#> 19   -1.1532416661      0
#> 20    1.1525671119      0
#> 21    0.1209121686      0
#> 22    0.8712855525      0
#> 23   -1.1069647854      0
#> 24    1.6553293434      0
#> 25    0.9092797985      0
#> 26    0.4743072216      0
#> 27    0.1443259761      0
#> 28   -0.9490552318      0
#> 29    0.7491553527      0
#> 30    0.4636749642      0
#> 31   -1.1629653487      0
#> 32    0.4524560291      0
#> 33   -0.3987774162      0
#> 34   -0.4816532065      0
#> 35    0.8173641516      0
#> 36    1.3511938388      0
#> 37   -1.6302885356      0
#> 38   -0.6756411549      0
#> 39    1.4913474547      0
#> 40    0.1717181144      0
#> 41   -0.6556602432      0
#> 42   -1.9569731067      0
#> 43    1.4268210139      0
#> 44   -0.3966098764      0
#> 45   -0.8447710750      0
#> 46   -1.0660503843      0
#> 47   -0.2662773566      0
#> 48    1.0825951941      0
#> 49   -0.5568959090      0
#> 50    0.9168076663      0
#> 51    0.4532172239      0
#> 52   -0.5302500778      0
#> 53    0.1077303524      0
#> 54   -0.1644073882      0
#> 55   -0.0966270472      0
#> 56   -0.6262875880      0
#> 57    0.3744645192      0
#> 58    0.9571630212      0
#> 59    0.6780085533      0
#> 60    0.1527178932      0
#> 61   -0.1031519446      0
#> 62   -1.6915673884      0
#> 63    0.1235872955      0
#> 64    0.9491331482      0
#> 65    1.3823608610      0
#> 66    1.1919823980      0
#> 67   -1.1007853503      0
#> 68   -1.2414693305      0
#> 69    0.6681043701      0
#> 70    1.9063780618      0
#> 71   -0.9811474853      0
#> 72    0.1096281754      0
#> 73    1.7897316180      0
#> 74    2.6292996937      0
#> 75    0.2880487135      0
#> 76   -0.5140027608      0
#> 77    0.9385777204      0
#> 78   -0.7354210955      0
#> 79   -0.9924193845      0
#> 80    1.8317154978      0
#> 81   -0.4038160474      0
#> 82    0.0630829475      0
#> 83   -0.2691362146      0
#> 84    0.0137046540      0
#> 85   -1.2743306053      0
#> 86   -2.0491076092      0
#> 87   -0.7588313860      0
#> 88    0.8841873567      0
#> 89    0.6210537676      0
#> 90    0.6161001675      0
#> 91   -0.3595416290      0
#> 92   -1.2160807807      0
#> 93   -0.3059883679      0
#> 94    0.2329435528      0
#> 95   -1.4166127211      0
#> 96    1.4487587869      0
#> 97    0.7358949537      0
#> 98   -0.3788559020      0
#> 99    0.0088664635      0
#> 100  -1.1741639277      0
#> 101   1.4640259877      0
#> 102   0.9395783052      0
#> 103  -0.0217683006      0
#> 104   0.8331227015      0
#> 105  -0.1080665278      0
#> 106  -0.8736436353      0
#> 107   0.3482486657      0
#> 108   0.5344938508      0
#> 109   0.3994941635      0
#> 110   0.4475506972      0
#> 111   1.3040240578      0
#> 112   0.1731718583      0
#> 113   0.9221849343      0
#> 114   0.1923774601      0
#> 115  -0.3439803355      0
#> 116  -0.8146275810      0
#> 117  -1.0011852048      0
#> 118   0.5820218157      0
#> 119   0.5260814402      0
#> 120  -0.2624117691      0
#> 121  -0.9659562499      0
#> 122   1.2949855459      0
#> 123  -0.3755373178      0
#> 124  -1.8017482440      0
#> 125   0.4460676784      0
#> 126   0.3135218508      0
#> 127   1.1670194132      0
#> 128  -0.2932768455      0
#> 129   0.2594545602      0
#> 130   0.9146186725      0
#> 131   0.9236855981      0
#> 132   0.7981494212      0
#> 133  -0.3448541023      0
#> 134   0.3772370406      0
#> 135   0.2422770790      0
#> 136  -1.6438455663      0
#> 137   0.4017062022      0
#> 138   0.8307795833      0
#> 139  -0.1618003394      0
#> 140   0.6068002179      0
#> 141  -0.5094409039      0
#> 142   1.4346578337      0
#> 143   1.5291430088      0
#> 144  -0.3829790797      0
#> 145   0.8206288262      0
#> 146   2.4927870528      0
#> 147   0.3633288797      0
#> 148  -1.4271791826      0
#> 149  -1.7462016325      0
#> 150   0.0795553074      0
#> 151   0.7090568442      0
#> 152   0.8273782976      0
#> 153   0.8391918795      0
#> 154   0.1470514385      0
#> 155   0.0313510343      0
#> 156   1.1692936707      0
#> 157   0.0593671799      0
#> 158   0.2081486277      0
#> 159  -0.0674860753      0
#> 160   1.6995606293      0
#> 161   0.1807870120      0
#> 162  -0.6329648038      0
#> 163  -0.1644250559      0
#> 164  -0.6070150778      0
#> 165  -1.4259434204      0
#> 166   0.6217102806      0
#> 167   0.0636817245      0
#> 168   0.0847321361      0
#> 169   0.6852526632      0
#> 170   1.7197804710      0
#> 171   0.0111257522      0
#> 172   0.4094646812      0
#> 173   1.1615230483      0
#> 174   1.1131571404      0
#> 175  -0.9270598871      0
#> 176  -1.0567528719      0
#> 177   1.4883154773      0
#> 178   0.2467860164      0
#> 179  -0.0938444056      0
#> 180   1.3816523136      0
#> 181   0.8232594037      0
#> 182  -0.0743830906      0
#> 183  -0.6286631126      0
#> 184   1.4396840677      0
#> 185   0.0181328650      0
#> 186  -2.0300729715      0
#> 187  -0.1001512843      0
#> 188   0.0933695401      0
#> 189   2.1982692780      0
#> 190   2.3808728499      0
#> 191  -0.6103269322      0
#> 192  -1.1153714260      0
#> 193   0.5392895217      0
#> 194  -0.0456833249      0
#> 195   0.8268673543      0
#> 196   0.0192768344      0
#> 197   2.0313079546      0
#> 198   1.1439840260      0
#> 199  -0.1293534395      0
#> 200   0.1489455992      0
#> 201  -1.0415793917      0
#> 202   1.0260853382      0
#> 203  -1.7864020312      0
#> 204  -0.0056259359      0
#> 205  -0.6695797197      0
#> 206   0.5072483214      0
#> 207   0.0215643604      0
#> 208  -1.4910014574      0
#> 209  -0.8632078493      0
#> 210   1.2782323934      0
#> 211  -1.6464272426      0
#> 212   0.4700041874      0
#> 213  -0.0684325538      0
#> 214   2.0369958447      0
#> 215  -0.7039182074      0
#> 216   1.3349149683      0
#> 217  -1.4734549794      0
#> 218  -0.7929696171      0
#> 219  -1.9876991250      0
#> 220  -1.1448719896      0
#> 221   0.7771629579      0
#> 222   0.3924666764      0
#> 223  -0.0459687780      0
#> 224  -0.6771740612      0
#> 225   1.3426410204      0
#> 226  -0.4252436911      0
#> 227  -0.2827631140      0
#> 228  -0.6989208255      0
#> 229  -1.3597211193      0
#> 230  -0.8794359954      0
#> 231  -1.1110613019      0
#> 232   0.8829447101      0
#> 233   0.2678973112      0
#> 234  -0.0331498918      0
#> 235   0.4379452506      0
#> 236   0.2126171550      0
#> 237  -0.5275351666      0
#> 238  -0.5676805078      0
#> 239   1.0766980959      0
#> 240  -0.4918932637      0
#> 241   0.0117522545      0
#> 242  -0.6869296497      0
#> 243  -1.1589553213      0
#> 244   0.7156311371      0
#> 245  -0.7102529696      0
#> 246   1.6332202026      0
#> 247   0.8146487983      0
#> 248   0.5862627445      0
#> 249   1.2006383131      0
#> 250  -1.1671251827      0
#> 251  -0.3913088195      0
#> 252   0.2958276747      0
#> 253   1.3841040155      0
#> 254   0.8175351341      0
#> 255  -0.0681079774      0
#> 256   1.3509631016      0
#> 257   1.7925377958      0
#> 258   1.4581515193      0
#> 259   0.6000014448      0
#> 260   0.2396952252      0
#> 261  -0.2304353554      0
#> 262   0.9329620693      0
#> 263   0.9163317841      0
#> 264   2.1736673122      0
#> 265  -0.2952985271      0
#> 266  -0.6332019292      0
#> 267   0.4326264254      0
#> 268   1.2596487563      0
#> 269  -0.2043552859      0
#> 270  -1.0365608763      0
#> 271   0.8188288730      0
#> 272   0.2356883900      0
#> 273   0.0788849695      0
#> 274  -0.3283068808      0
#> 275  -0.9690683624      0
#> 276   0.3529070817      0
#> 277   1.1784627135      0
#> 278   1.9685726285      0
#> 279   2.2207155899      0
#> 280  -0.0486505096      0
#> 281   2.5154209533      0
#> 282   0.7003329849      0
#> 283  -0.6977954774      0
#> 284  -1.0173694305      0
#> 285  -0.3385323717      0
#> 286   0.9686548003      0
#> 287  -0.4613939939      0
#> 288   1.1546839446      0
#> 289  -0.5316967837      0
#> 290  -0.9113729821      0
#> 291  -0.8280012562      0
#> 292  -0.3710023421      0
#> 293  -0.6892330909      0
#> 294   0.1580699909      0
#> 295  -0.5741198208      0
#> 296  -0.8755472891      0
#> 297  -0.4610817486      0
#> 298  -0.3888221859      0
#> 299  -0.5495561456      0
#> 300   0.2933898653      0
#> 301  -0.0764002524      0
#> 302   1.4586243743      0
#> 303   0.3784306515      0
#> 304   0.4823812285      0
#> 305   1.9418978515      0
#> 306  -0.1000335956      0
#> 307   2.0238010886      0
#> 308   0.7141527738      0
#> 309  -0.0736464053      0
#> 310   0.0561679535      0
#> 311   0.8860420260      0
#> 312  -0.5783950937      0
#> 313  -0.2690929351      0
#> 314   0.3032215616      0
#> 315  -0.6885327081      0
#> 316   0.3866064897      0
#> 317   0.1733919686      0
#> 318   1.8942484096      0
#> 319   2.1044040178      0
#> 320  -0.0371951307      0
#> 321   0.1522476480      0
#> 322   1.1266958383      0
#> 323  -0.4554209711      0
#> 324   0.9829454019      0
#> 325   2.6889090739      0
#> 326   0.3956701279      0
#> 327  -1.0551584958      0
#> 328   0.2188109544      0
#> 329  -0.5144313019      0
#> 330  -1.1144153138      0
#> 331  -0.5440031901      0
#> 332   0.9687338355      0
#> 333  -0.6536832879      0
#> 334  -0.3860546831      0
#> 335   1.1375981351      0
#> 336   0.1473278389      0
#> 337  -0.3606951982      0
#> 338  -0.4307231305      0
#> 339   1.0712221942      0
#> 340  -4.0853558946      0
#> 341   0.4804548871      0
#> 342   0.4036467591      0
#> 343  -0.2983115936      0
#> 344  -1.3499021492      0
#> 345  -1.0745074386      0
#> 346   0.2992638179      0
#> 347  -0.4479006275      0
#> 348  -0.0818823539      0
#> 349  -0.0078389103      0
#> 350  -0.1821647973      0
#> 351  -0.1395860238      0
#> 352  -2.0571977431      0
#> 353   0.3341085103      0
#> 354  -0.7319362666      0
#> 355   0.1948012757      0
#> 356  -0.4662454440      0
#> 357   2.1877542023      0
#> 358  -0.4484594110      0
#> 359  -0.7731157528      0
#> 360   1.3411247635      0
#> 361  -0.0218519380      0
#> 362   0.5124510731      0
#> 363  -0.0762928046      0
#> 364  -0.2440105617      0
#> 365  -0.8403691567      0
#> 366   0.9568027463      0
#> 367   0.3644538927      0
#> 368   0.5708072653      0
#> 369   3.2229837030      0
#> 370   1.6182454343      0
#> 371  -0.4808830776      0
#> 372  -0.9239003883      0
#> 373   0.7297787251      0
#> 374  -2.3304593690      0
#> 375   1.6183484311      0
#> 376   1.9125809216      0
#> 377  -2.0178477356      0
#> 378   0.0718576625      0
#> 379  -1.2916825407      0
#> 380  -0.3503240800      0
#> 381   0.2435804444      0
#> 382   0.2558543597      0
#> 383   0.1200097120      0
#> 384  -1.0947723365      0
#> 385   0.6904386869      0
#> 386   0.7664067851      0
#> 387   0.3279843157      0
#> 388   0.5993085389      0
#> 389   1.6943204582      0
#> 390   0.3003483485      0
#> 391  -0.9727095877      0
#> 392  -2.1717591391      0
#> 393  -0.8809510760      0
#> 394  -1.7753676559      0
#> 395   0.8142816962      0
#> 396   0.8937955078      0
#> 397   0.4727221169      0
#> 398   2.0790271529      0
#> 399  -1.8300731803      0
#> 400   0.4600240771      0
#> 401   1.3140396426      0
#> 402   0.6647733479      0
#> 403  -1.3860009690      0
#> 404   0.8858038274      0
#> 405  -0.5027371873      0
#> 406   0.7731289566      0
#> 407  -0.2203435448      0
#> 408  -0.2386472072      0
#> 409   0.3420702618      0
#> 410   1.4988218724      0
#> 411   1.4310338557      0
#> 412  -1.2875249666      0
#> 413   1.5165657555      0
#> 414   0.8976352802      0
#> 415  -0.5018058248      0
#> 416   0.8624995783      0
#> 417  -1.5616707477      0
#> 418   0.3189282500      0
#> 419   0.0292565185      0
#> 420   1.9268697709      0
#> 421   2.0467315583      0
#> 422  -0.3022359307      0
#> 423   0.2669052979      0
#> 424  -0.1832126889      0
#> 425   0.8434545693      0
#> 426   0.2217565685      0
#> 427   0.6334126578      0
#> 428  -1.5592643843      0
#> 429  -1.2604378592      0
#> 430   0.1929033081      0
#> 431   1.1684166107      0
#> 432   1.1313741378      0
#> 433  -0.0109090254      0
#> 434  -1.0698153758      0
#> 435   0.0481473886      0
#> 436  -0.9350410191      0
#> 437   0.7623011704      0
#> 438   1.4111706810      0
#> 439  -0.4502169245      0
#> 440   0.1413584139      0
#> 441  -0.9489267103      0
#> 442  -0.2694044014      0
#> 443   0.0553059786      0
#> 444  -0.1054594738      0
#> 445   0.2877749411      0
#> 446  -0.2581764703      0
#> 447  -0.8832122066      0
#> 448  -0.2509238322      0
#> 449   0.6533816792      0
#> 450  -0.3698393077      0
#> 451   0.0557385255      0
#> 452   0.9187558539      0
#> 453   0.2989299046      0
#> 454   0.7460651635      0
#> 455   0.9875628447      0
#> 456   0.9448000760      0
#> 457   0.6800433780      0
#> 458   1.4957202885      0
#> 459   0.0129599516      0
#> 460   0.1909685622      0
#> 461  -0.0212620588      0
#> 462  -1.6468451497      0
#> 463  -0.3681811941      0
#> 464   0.1235331809      0
#> 465  -0.0669045859      0
#> 466   0.2327595136      0
#> 467  -1.4184303121      0
#> 468  -0.6281409851      0
#> 469   1.7076797833      0
#> 470   0.0956711865      0
#> 471   0.1119598280      0
#> 472  -0.8586910821      0
#> 473  -1.5605947097      0
#> 474  -1.2128385489      0
#> 475   0.1757322715      0
#> 476  -0.2341358643      0
#> 477  -0.2091756127      0
#> 478   1.9275408022      0
#> 479  -0.4916899891      0
#> 480   0.5548171723      0
#> 481  -0.5689772611      0
#> 482   1.0390506790      0
#> 483   0.3430341600      0
#> 484   0.1494801814      0
#> 485   1.6829499398      0
#> 486   0.9089350425      0
#> 487   2.0668545118      0
#> 488   2.0707445474      0
#> 489  -0.4432385883      0
#> 490  -1.2510684520      0
#> 491  -1.2891133785      0
#> 492   2.1692867433      0
#> 493  -0.5311100238      0
#> 494   0.8054093847      0
#> 495  -0.5392987331      0
#> 496   0.7098992635      0
#> 497  -2.9034932633      0
#> 498  -0.0752592914      0
#> 499  -1.5649693941      0
#> 500  -0.6758038249      0
#> 501  -1.6694202647      0
#> 502   0.2460647083      0
#> 503   1.5651211605      0
#> 504  -1.8944629457      0
#> 505  -1.0415373240      0
#> 506   0.2771551828      0
#> 507  -0.4086705441      0
#> 508   0.7314711115      0
#> 509   1.1210640337      0
#> 510   0.1241829921      0
#> 511  -0.5099018795      0
#> 512   0.5389253238      0
#> 513   3.0569084899      0
#> 514  -0.3141473030      0
#> 515   0.1883687606      0
#> 516  -0.5425741888      0
#> 517  -0.8256338397      0
#> 518   0.4624787792      0
#> 519  -0.2644775746      0
#> 520   0.2252345580      0
#> 521   0.8115850023      0
#> 522  -0.9867393900      0
#> 523   0.7002516323      0
#> 524  -0.0121206481      0
#> 525   0.7718376112      0
#> 526   1.1785042714      0
#> 527   0.6731186095      0
#> 528   1.4045450143      0
#> 529   0.8663392663      0
#> 530  -0.0302473885      0
#> 531   0.3556683634      0
#> 532   0.0349002564      0
#> 533   0.7628836628      0
#> 534  -0.0110215420      0
#> 535  -0.0768846612      0
#> 536  -0.4230001639      0
#> 537   1.1101731046      0
#> 538  -0.8070970609      0
#> 539  -0.1068618679      0
#> 540  -1.6035732961      0
#> 541  -0.2208048810      0
#> 542  -1.8078348147      0
#> 543   0.7378544374      0
#> 544   0.1980916127      0
#> 545  -1.0409942952      0
#> 546   0.0863916641      0
#> 547   0.1540918790      0
#> 548  -0.3652995535      0
#> 549  -1.3806096302      0
#> 550   2.3756259184      0
#> 551   0.9869485303      0
#> 552  -0.5652432510      0
#> 553   0.6640894778      0
#> 554   1.0583603422      0
#> 555  -0.5750011383      0
#> 556   0.1471818263      0
#> 557   0.0076119909      0
#> 558   2.2094112389      0
#> 559   0.2197687110      0
#> 560   0.1468223710      0
#> 561  -0.6890051057      0
#> 562  -0.8461997939      0
#> 563  -2.3784694638      0
#> 564  -0.2267848405      0
#> 565  -1.2632226730      0
#> 566   1.5160876193      0
#> 567  -0.6702008767      0
#> 568   1.0472100653      0
#> 569   0.0125412057      0
#> 570   1.5731840736      0
#> 571   2.2273619913      0
#> 572  -1.6453949836      0
#> 573  -0.4446752760      0
#> 574   0.9002561497      0
#> 575  -0.9670419097      0
#> 576   0.1133021981      0
#> 577   1.1691531034      0
#> 578   0.7951401397      0
#> 579  -1.2511776231      0
#> 580   0.0750660724      0
#> 581  -0.1992016188      0
#> 582   0.3839948148      0
#> 583   1.5040829657      0
#> 584   0.3768370872      0
#> 585   0.1305770863      0
#> 586   0.7433158838      0
#> 587  -0.9981267691      0
#> 588  -0.8495055905      0
#> 589  -0.2635418378      0
#> 590   1.3728663077      0
#> 591   0.0907441087      0
#> 592  -0.2101110789      0
#> 593  -0.9344792560      0
#> 594  -1.0469852646      0
#> 595   0.7324174691      0
#> 596   1.7513631180      0
#> 597  -0.1859035950      0
#> 598   0.1610012132      0
#> 599  -0.1676776338      0
#> 600  -0.7871430951      0
#> 601  -0.3046816052      0
#> 602  -1.1862544201      0
#> 603   1.2219280342      0
#> 604   0.2177574718      0
#> 605   0.2830987235      0
#> 606   0.7498660807      0
#> 607  -0.5039270360      0
#> 608   2.2318256676      0
#> 609  -0.7442943961      0
#> 610   0.7902206668      0
#> 611   0.9129990626      0
#> 612  -0.2103467418      0
#> 613   0.6937412945      0
#> 614   0.2859805645      0
#> 615  -1.2679664445      0
#> 616  -1.1696981379      0
#> 617  -0.1927274763      0
#> 618  -0.3442729771      0
#> 619   0.0362376832      0
#> 620  -0.3790070266      0
#> 621  -0.3749671739      0
#> 622  -1.2609298305      0
#> 623   0.6643429067      0
#> 624  -0.4257373775      0
#> 625  -1.1795959601      0
#> 626  -0.5692325083      0
#> 627  -1.1034143721      0
#> 628   0.0329077781      0
#> 629  -0.1481980671      0
#> 630  -1.2493991088      0
#> 631  -2.0535329974      0
#> 632   0.8048736475      0
#> 633  -2.3775548936      0
#> 634  -1.7733120867      0
#> 635  -1.3178191769      0
#> 636   2.3960653594      0
#> 637   0.0598745425      0
#> 638   0.1572497726      0
#> 639  -0.3329210750      0
#> 640   0.7267580230      0
#> 641   2.0177338540      0
#> 642  -0.1937477581      0
#> 643   0.8046580954      0
#> 644   0.1565007276      0
#> 645  -1.2413861282      0
#> 646   0.0297959887      0
#> 647  -1.3109296848      0
#> 648  -0.3649254521      0
#> 649  -2.3088012441      0
#> 650  -0.6747540242      0
#> 651   0.9153633559      0
#> 652  -0.0485366999      0
#> 653   0.3644991091      0
#> 654   2.1142072091      0
#> 655  -1.2453214543      0
#> 656  -0.1035361550      0
#> 657  -0.7468978400      0
#> 658   0.3151182184      0
#> 659   0.4456129789      0
#> 660   0.7211192631      0
#> 661   0.5762747117      0
#> 662  -0.1414702382      0
#> 663   1.5420777673      0
#> 664  -0.4381407246      0
#> 665   0.1120230551      0
#> 666  -0.0497783496      0
#> 667  -1.8079019123      0
#> 668   0.5473444800      0
#> 669  -0.1638474266      0
#> 670  -0.8171403973      0
#> 671   0.5868940580      0
#> 672  -0.3212539080      0
#> 673   1.2597458721      0
#> 674  -1.3894026622      0
#> 675  -0.4886161014      0
#> 676  -1.5569234417      0
#> 677  -0.1165963847      0
#> 678   0.6421508248      0
#> 679  -0.1922297438      0
#> 680   0.3769739974      0
#> 681  -0.3277774855      0
#> 682  -0.9255086768      0
#> 683   0.5559629649      0
#> 684   0.1737748853      0
#> 685   0.0460165070      0
#> 686  -0.5025606707      0
#> 687  -0.8371563939      0
#> 688  -0.1980296066      0
#> 689   1.5451185752      0
#> 690   0.0338798619      0
#> 691  -1.1772363216      0
#> 692   1.0064823607      0
#> 693  -0.3966399148      0
#> 694   0.6258008629      0
#> 695  -1.1395897165      0
#> 696  -0.2132621408      0
#> 697  -1.2126016579      0
#> 698  -0.2804875518      0
#> 699   0.4482528162      0
#> 700   1.1120265557      0
#> 701  -0.4066856666      0
#> 702  -1.9253580604      0
#> 703  -1.2469008666      0
#> 704   0.2960682136      0
#> 705  -0.6677898308      0
#> 706  -0.6045172085      0
#> 707  -2.1911098158      0
#> 708  -0.3496619220      0
#> 709  -0.8226082166      0
#> 710  -0.4717247362      0
#> 711   0.7587851749      0
#> 712  -0.2983609168      0
#> 713  -0.7863660475      0
#> 714  -1.2090668097      0
#> 715   0.5425900029      0
#> 716  -0.7937761941      0
#> 717  -0.7590818096      0
#> 718  -1.6456940952      0
#> 719   0.6615110474      0
#> 720   2.2317980356      0
#> 721   0.3035924015      0
#> 722  -0.3733309847      0
#> 723  -1.4215210571      0
#> 724   0.0274754862      0
#> 725   0.6747470681      0
#> 726  -1.1032077119      0
#> 727  -0.0898148661      0
#> 728  -0.4475633493      0
#> 729  -1.3017029657      0
#> 730  -0.0345332578      0
#> 731  -0.0829496133      0
#> 732   0.8721862255      0
#> 733   0.1647417052      0
#> 734  -0.9181775508      0
#> 735   1.1198111068      0
#> 736   0.5971037441      0
#> 737  -0.8836011990      0
#> 738  -1.1106156887      0
#> 739  -0.0690422047      0
#> 740   0.2476630208      0
#> 741   0.1219286077      0
#> 742  -1.1121452235      0
#> 743  -0.4866359059      0
#> 744   1.7524822815      0
#> 745  -0.9830636227      0
#> 746  -0.2721371368      0
#> 747  -0.9316886927      0
#> 748   1.1712194136      0
#> 749   0.2774827328      0
#> 750   0.3369391695      0
#> 751   0.2519527189      0
#> 752   0.9786008356      0
#> 753  -0.6070229756      0
#> 754   0.7164160648      0
#> 755  -0.9205271914      0
#> 756  -0.6953870064      0
#> 757  -0.6869121256      0
#> 758  -0.5447305225      0
#> 759   1.6495634928      0
#> 760   0.0776067495      0
#> 761  -0.9722121405      0
#> 762   0.9724168084      0
#> 763  -0.2207708569      0
#> 764   0.6064538077      0
#> 765  -0.4902949325      0
#> 766   0.2406319657      0
#> 767  -0.0121921158      0
#> 768  -0.8039641538      0
#> 769   0.6587483290      0
#> 770   1.9964435195      0
#> 771   1.0850012215      0
#> 772   0.2076817748      0
#> 773  -0.0399654637      0
#> 774  -0.0475636863      0
#> 775   0.9559826412      0
#> 776  -0.1985387461      0
#> 777   0.5618181785      0
#> 778   1.0650181981      0
#> 779  -0.1170106771      0
#> 780   2.1111285823      0
#> 781  -1.2057942329      0
#> 782  -0.8469840935      0
#> 783   1.2496773431      0
#> 784   0.6527923160      0
#> 785  -0.2154859189      0
#> 786  -0.8186692332      0
#> 787   0.1295949202      0
#> 788  -0.3134374618      0
#> 789   0.0710974293      0
#> 790   0.5260031718      0
#> 791   1.5708331341      0
#> 792   1.1989238705      0
#> 793   0.2711568479      0
#> 794  -0.2477473837      0
#> 795   0.6447503388      0
#> 796   0.0438201658      0
#> 797   0.9257081898      0
#> 798  -0.5462088470      0
#> 799   0.0855414014      0
#> 800   0.4534602876      0
#> 801   0.3960542597      0
#> 802   0.4940654221      0
#> 803   0.4717114514      0
#> 804  -0.2844704904      0
#> 805   0.6935413838      0
#> 806  -0.4302109846      0
#> 807  -1.1140328095      0
#> 808  -2.1198265153      0
#> 809  -0.1110757146      0
#> 810  -1.4011630151      0
#> 811  -1.5237849626      0
#> 812  -1.4541927968      0
#> 813  -0.6711502601      0
#> 814  -0.9764228336      0
#> 815  -0.1908526930      0
#> 816  -1.7850222414      0
#> 817  -2.3124378589      0
#> 818   0.1364257280      0
#> 819  -1.3347716914      0
#> 820  -0.2429348256      0
#> 821  -1.8366620503      0
#> 822  -0.0964942121      0
#> 823  -1.1451581287      0
#> 824  -0.4238117061      0
#> 825   0.7646549174      0
#> 826  -0.1590769531      0
#> 827   1.2757252481      0
#> 828   0.1379739499      0
#> 829   0.0066868742      0
#> 830  -0.0242052614      0
#> 831   0.3197999925      0
#> 832   0.0201421751      0
#> 833  -0.3682855333      0
#> 834   0.1057953979      0
#> 835   0.4079606856      0
#> 836   1.7716360622      0
#> 837   1.1952164103      0
#> 838  -0.2654524159      0
#> 839  -0.4127821285      0
#> 840  -1.3475308704      0
#> 841   0.6340580143      0
#> 842   0.4921464075      0
#> 843  -0.8395207674      0
#> 844  -0.4475298203      0
#> 845   0.1028658253      0
#> 846   0.6048820028      0
#> 847   0.0934114659      0
#> 848  -0.9636906025      0
#> 849  -1.2744519237      0
#> 850   0.5096358714      0
#> 851   1.3175053790      0
#> 852  -0.7124864993      0
#> 853  -0.5697774411      0
#> 854  -0.3150281090      0
#> 855  -0.2131289059      0
#> 856  -1.0964389816      0
#> 857   0.3212144242      0
#> 858  -1.2673661047      0
#> 859   0.3557233446      0
#> 860  -0.1584682996      0
#> 861  -0.0255461634      0
#> 862   0.8980073091      0
#> 863   0.3101330709      0
#> 864  -0.3901371731      0
#> 865  -0.4484031809      0
#> 866  -0.0004438596      0
#> 867   0.2495988348      0
#> 868  -1.8229278878      0
#> 869   0.4923130215      0
#> 870  -0.1697857705      0
#> 871  -0.0962768815      0
#> 872   0.5685151432      0
#> 873  -0.4395841649      0
#> 874  -0.8413534394      0
#> 875  -0.8486969103      0
#> 876   0.5821150094      0
#> 877   0.8703360419      0
#> 878  -0.2540837962      0
#> 879   1.8290155838      1
#> 880   0.6628355465      1
#> 881   0.0951833515      1
#> 882   0.8024061631      1
#> 883   0.3670409332      1
#> 884   0.1346599438      1
#> 885   1.3425791457      1
#> 886   1.2936229633      1
#> 887   1.9608289657      1
#> 888   0.7784527674      1
#> 889   2.5757512469      1
#> 890   0.2397289396      1
#> 891   0.6688873703      1
#> 892   2.5339239890      1
#> 893   1.9203496129      1
#> 894   1.0468534065      1
#> 895  -0.4300955806      1
#> 896   0.4613517454      1
#> 897   0.1615477863      1
#> 898   0.1914002887      1
#> 899   1.8823681710      1
#> 900  -0.0709011047      1
#> 901   0.7511249899      1
#> 902   1.3278265467      1
#> 903   1.6544019618      1
#> 904   0.0098245585      1
#> 905   0.7423632328      1
#> 906   1.9549968034      1
#> 907   0.1124649017      1
#> 908   0.2826798961      1
#> 909  -0.8900792410      1
#> 910   3.7352024988      1
#> 911   0.6787829732      1
#> 912   1.8500409014      1
#> 913   0.2351437501      1
#> 914  -0.4251293572      1
#> 915   0.0092622176      1
#> 916   0.3585571657      1
#> 917   0.3709626490      1
#> 918  -0.9202024183      1
#> 919   0.0152317917      1
#> 920   0.3238967814      1
#> 921   0.2837880751      1
#> 922  -0.0298122214      1
#> 923   1.2147406992      1
#> 924   0.5320188862      1
#> 925   2.5210416922      1
#> 926   0.0566260913      1
#> 927   0.4511477240      1
#> 928   0.6343702765      1
#> 929  -1.2058993047      1
#> 930  -0.8507199249      1
#> 931   2.4117426301      1
#> 932   0.4058543097      1
#> 933   0.9338562531      1
#> 934   0.3977088495      1
#> 935   0.0531941473      1
#> 936   1.7710547595      1
#> 937   0.3533496941      1
#> 938   1.6071143448      1
#> 939   0.9447502900      1
#> 940   2.1754952082      1
#> 941   1.7036207168      1
#> 942   2.8605378252      1
#> 943   0.0833982961      1
#> 944   2.4429114813      1
#> 945   1.6227681219      1
#> 946   0.6344051070      1
#> 947   0.0110720619      1
#> 948   0.9459379648      1
#> 949  -0.2207797919      1
#> 950  -0.6268882114      1
#> 951   1.7941350391      1
#> 952   2.6232978324      1
#> 953   1.2336439499      1
#> 954   1.4306086213      1
#> 955   1.0897868610      1
#> 956   1.8339571185      1
#> 957   1.5252211531      1
#> 958   0.3977168634      1
#> 959   0.8984843241      1
#> 960   2.2526427691      1
#> 961  -0.8208038615      1
#> 962   1.3444314797      1
#> 963  -0.7182960091      1
#> 964   2.5534949820      1
#> 965   1.3084137477      1
#> 966   0.1539763291      1
#> 967   1.3499324206      1
#> 968   1.3971710176      1
#> 969   1.1199613973      1
#> 970   0.3037397946      1
#> 971   0.6205599562      1
#> 972   1.6399897057      1
#> 973   2.3192787017      1
#> 974  -0.1020933774      1
#> 975   0.6339828710      1
#> 976   1.8483500657      1
#> 977   0.5789721235      1
#> 978   0.6858048816      1
#> 979   2.0336670805      1
#> 980   0.8243188061      1
#> 981   1.6173958422      1
#> 982   0.3146291406      1
#> 983   2.1927658418      1
#> 984  -1.3891780170      1
#> 985   1.1122559109      1
#> 986   0.2451361646      1
#> 987  -0.9275201121      1
#> 988   2.3775812788      1
#> 989   0.5001245244      1
#> 990   1.7988240133      1
#> 991   1.6909952179      1
#> 992  -0.0675085524      1
#> 993  -0.4205470253      1
#> 994  -0.0675268123      1
#> 995   0.3730400957      1
#> 996   1.3050155533      1
#> 997   1.0985008853      1
#> 998   0.0429171493      1
#> 999   0.0057351095      1
#> 1000  1.8304908423      1