Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     1.454673806      0
#> 2    -0.035141995      0
#> 3     0.200134710      0
#> 4     1.555174846      0
#> 5     0.456463974      0
#> 6     0.518716017      0
#> 7     0.327441065      0
#> 8    -0.364146717      0
#> 9     0.196661678      0
#> 10   -0.833568285      0
#> 11   -0.902909347      0
#> 12   -0.326250729      0
#> 13   -0.329850749      0
#> 14    1.933649019      0
#> 15    0.987090296      0
#> 16    1.949509155      0
#> 17    0.568704489      0
#> 18    0.955518736      0
#> 19   -0.167586962      0
#> 20   -0.954498683      0
#> 21    1.756307168      0
#> 22    1.470576127      0
#> 23    0.087498026      0
#> 24    0.587722890      0
#> 25    0.920929503      0
#> 26    1.626289338      0
#> 27   -0.063539858      0
#> 28   -0.164588930      0
#> 29    1.131118586      0
#> 30   -0.244037722      0
#> 31    0.218144914      0
#> 32   -0.007927860      0
#> 33   -1.118064269      0
#> 34    0.780328129      0
#> 35   -0.402530336      0
#> 36   -1.514643924      0
#> 37   -0.671820774      0
#> 38    0.028652708      0
#> 39   -1.143251255      0
#> 40    0.161717512      0
#> 41   -1.113166612      0
#> 42    0.446633817      0
#> 43    1.227602302      0
#> 44   -0.044749416      0
#> 45   -0.650013141      0
#> 46    0.186268727      0
#> 47   -0.395114176      0
#> 48   -2.069479622      0
#> 49   -2.334956584      0
#> 50    0.118356522      0
#> 51    1.049469294      0
#> 52   -1.790007596      0
#> 53   -0.356336045      0
#> 54    0.148658810      0
#> 55    1.044676558      0
#> 56   -0.401321803      0
#> 57   -0.641949606      0
#> 58   -1.108499470      0
#> 59    0.441963044      0
#> 60    0.892022007      0
#> 61    1.909351706      0
#> 62    0.674227467      0
#> 63    0.731700287      0
#> 64   -0.409151456      0
#> 65    0.781824617      0
#> 66    0.537711658      0
#> 67   -0.278080743      0
#> 68   -1.342157033      0
#> 69    0.048671392      0
#> 70   -0.183189511      0
#> 71   -0.459319412      0
#> 72   -1.806312869      0
#> 73    0.116158257      0
#> 74   -0.770303823      0
#> 75    1.748659095      0
#> 76    1.655593254      0
#> 77   -0.700282534      0
#> 78    1.362562114      0
#> 79   -1.051474891      0
#> 80    0.893743490      0
#> 81   -1.083264757      0
#> 82    0.375765967      0
#> 83   -2.447585305      0
#> 84   -0.060912635      0
#> 85    0.231552571      0
#> 86    0.269163853      0
#> 87   -0.056773061      0
#> 88    0.578000586      0
#> 89   -0.523675229      0
#> 90    2.463194044      0
#> 91    0.651714888      0
#> 92   -0.745065823      0
#> 93   -0.702055158      0
#> 94   -1.042968086      0
#> 95   -0.144127938      0
#> 96    0.831463370      0
#> 97    0.326568027      0
#> 98    2.678638302      0
#> 99   -1.396019210      0
#> 100  -0.411401431      0
#> 101   0.743675941      0
#> 102  -0.036704834      0
#> 103  -0.269580668      0
#> 104   1.647082541      0
#> 105  -0.691387442      0
#> 106  -0.441902921      0
#> 107   0.630847988      0
#> 108  -0.760884373      0
#> 109   1.082983054      0
#> 110  -1.608565214      0
#> 111  -1.055220046      0
#> 112  -0.392626120      0
#> 113  -0.407221389      0
#> 114   0.646297218      0
#> 115   0.061881430      0
#> 116   1.273082906      0
#> 117   1.097132877      0
#> 118  -0.291233196      0
#> 119  -0.652657363      0
#> 120   1.072762492      0
#> 121  -0.414591574      0
#> 122   0.154226989      0
#> 123   3.313994498      0
#> 124  -0.057999739      0
#> 125   0.613289094      0
#> 126   1.003568761      0
#> 127   0.489923924      0
#> 128   1.133085148      0
#> 129  -0.540131212      0
#> 130  -0.440086544      0
#> 131   1.255615222      0
#> 132   0.379865022      0
#> 133  -1.730577091      0
#> 134   0.463144187      0
#> 135  -0.610716514      0
#> 136   0.707681819      0
#> 137   0.005281549      0
#> 138  -1.896165792      0
#> 139   0.658077844      0
#> 140  -1.978878949      0
#> 141  -0.115465364      0
#> 142  -0.137501478      0
#> 143  -0.512123211      0
#> 144   0.823887383      0
#> 145  -1.178279398      0
#> 146   0.343385508      0
#> 147   0.376734397      0
#> 148  -0.093784680      0
#> 149  -0.021466111      0
#> 150  -1.797103690      0
#> 151   1.088000906      0
#> 152   2.141793984      0
#> 153   0.553316927      0
#> 154   1.069079096      0
#> 155   0.682903065      0
#> 156   0.912316298      0
#> 157  -1.126747316      0
#> 158   0.872368711      0
#> 159   0.660755903      0
#> 160   0.338262266      0
#> 161   0.302110181      0
#> 162   0.458869377      0
#> 163   0.364418624      0
#> 164   1.021978063      0
#> 165  -0.516948307      0
#> 166   0.210246438      0
#> 167   0.606948185      0
#> 168   2.075266289      0
#> 169   1.571679915      0
#> 170   1.156193262      0
#> 171  -0.721090035      0
#> 172  -0.310261385      0
#> 173  -1.292132162      0
#> 174  -0.755296542      0
#> 175   0.287390807      0
#> 176   1.582790528      0
#> 177  -0.432633950      0
#> 178   0.165052454      0
#> 179   2.019091573      0
#> 180   1.422846394      0
#> 181  -0.005413869      0
#> 182  -0.271944196      0
#> 183  -0.480355781      0
#> 184  -0.680469102      0
#> 185  -0.490923689      0
#> 186  -0.468720755      0
#> 187  -0.446582704      0
#> 188   0.066247989      0
#> 189  -1.238804165      0
#> 190   0.172056158      0
#> 191  -0.738845618      0
#> 192   0.562863384      0
#> 193  -0.542062533      0
#> 194   0.688151260      0
#> 195   0.589664563      0
#> 196  -0.192454741      0
#> 197   1.228461648      0
#> 198   1.223444036      0
#> 199   0.024551467      0
#> 200   0.577224226      0
#> 201   0.029589036      0
#> 202   0.254774541      0
#> 203  -0.107298317      0
#> 204  -0.538007864      0
#> 205  -0.865384486      0
#> 206   0.288537219      0
#> 207  -0.442427895      0
#> 208   0.354278405      0
#> 209  -0.369830955      0
#> 210   0.415949973      0
#> 211   0.885947786      0
#> 212  -1.060590479      0
#> 213  -1.879393031      0
#> 214   0.733537602      0
#> 215   0.081760899      0
#> 216  -1.269287669      0
#> 217  -0.253961365      0
#> 218   0.324745169      0
#> 219  -2.202307516      0
#> 220   1.482375206      0
#> 221  -0.221721933      0
#> 222   0.306507973      0
#> 223  -0.288146150      0
#> 224  -0.319945503      0
#> 225  -1.145081730      0
#> 226  -2.279135135      0
#> 227  -0.386753456      0
#> 228  -0.732565392      0
#> 229  -0.856905611      0
#> 230   1.763867576      0
#> 231  -0.133846935      0
#> 232  -0.328429109      0
#> 233  -1.112146693      0
#> 234  -0.965602058      0
#> 235  -0.266499162      0
#> 236   0.529228745      0
#> 237   1.193799216      0
#> 238  -1.913827397      0
#> 239   0.669817387      0
#> 240  -0.216813075      0
#> 241  -0.730573635      0
#> 242   0.772680301      0
#> 243   0.230352196      0
#> 244  -1.340697670      0
#> 245  -0.790280140      0
#> 246   0.934696290      0
#> 247  -1.284511272      0
#> 248  -0.158870388      0
#> 249  -0.907151746      0
#> 250   1.859779801      0
#> 251   0.564481033      0
#> 252  -0.274293209      0
#> 253  -2.262879826      0
#> 254  -0.845211818      0
#> 255   1.606144128      0
#> 256   2.462016020      0
#> 257   0.071604163      0
#> 258   0.746579211      0
#> 259   1.574906369      0
#> 260   0.796310545      0
#> 261  -0.144847817      0
#> 262   1.443025442      0
#> 263   0.634988090      0
#> 264  -1.271730397      0
#> 265  -1.423968023      0
#> 266   0.137002061      0
#> 267  -0.143659402      0
#> 268   0.644766794      0
#> 269   0.373427307      0
#> 270  -1.118404421      0
#> 271   0.056634082      0
#> 272   0.373220908      0
#> 273  -1.099038822      0
#> 274   0.436442838      0
#> 275  -0.905217480      0
#> 276  -0.453501395      0
#> 277   0.344516508      0
#> 278  -2.144378515      0
#> 279   0.385520693      0
#> 280  -0.030564064      0
#> 281  -0.908458072      0
#> 282   1.328435961      0
#> 283   0.608130027      0
#> 284   0.056839625      0
#> 285   0.995790761      0
#> 286   1.398597295      0
#> 287  -0.470076989      0
#> 288  -2.121276704      0
#> 289   0.122185982      0
#> 290  -0.225904902      0
#> 291   0.900568514      0
#> 292  -1.555477187      0
#> 293  -1.302480672      0
#> 294  -1.393986128      0
#> 295   0.176065238      0
#> 296  -0.981232935      0
#> 297  -0.423738859      0
#> 298   0.341432315      0
#> 299  -0.372631923      0
#> 300  -0.420834962      0
#> 301   0.041434725      0
#> 302   0.862557976      0
#> 303   1.070732776      0
#> 304  -3.409722433      0
#> 305   1.473981083      0
#> 306  -0.228040482      0
#> 307  -1.048240828      0
#> 308   1.068869663      0
#> 309  -0.092391491      0
#> 310   1.021569095      0
#> 311  -0.601988711      0
#> 312  -1.177178418      0
#> 313   0.399926614      0
#> 314  -0.666793143      0
#> 315  -1.290603149      0
#> 316  -0.988747462      0
#> 317   0.467676409      0
#> 318   0.500716857      0
#> 319  -1.474197702      0
#> 320  -0.473603761      0
#> 321  -1.002189149      0
#> 322  -1.203189639      0
#> 323  -1.165642989      0
#> 324  -0.197963568      0
#> 325  -1.113486109      0
#> 326  -1.552962330      0
#> 327   1.969337675      0
#> 328  -0.251609105      0
#> 329  -0.045551123      0
#> 330   1.135256983      0
#> 331  -0.508523638      0
#> 332   0.991519972      0
#> 333  -1.703743334      0
#> 334   0.855882998      0
#> 335  -0.132056344      0
#> 336  -1.348947815      0
#> 337  -1.112334840      0
#> 338  -0.649787402      0
#> 339  -1.250510564      0
#> 340   0.375637858      0
#> 341  -0.903289288      0
#> 342  -0.805421495      0
#> 343   0.350793942      0
#> 344  -0.206017122      0
#> 345  -0.959061201      0
#> 346  -0.543776932      0
#> 347  -0.619816748      0
#> 348  -0.815877494      0
#> 349  -1.617556849      0
#> 350   0.815227380      0
#> 351   0.726506641      0
#> 352   0.036340593      0
#> 353   0.098196135      0
#> 354   0.635893049      0
#> 355  -0.406174966      0
#> 356   0.089360773      0
#> 357  -1.190365723      0
#> 358  -0.887870154      0
#> 359  -0.532867479      0
#> 360  -1.347028787      0
#> 361  -1.034464306      0
#> 362  -0.064700118      0
#> 363   0.202461693      0
#> 364  -0.864357335      0
#> 365  -1.852843637      0
#> 366   0.341296675      0
#> 367   0.087484729      0
#> 368  -0.917604164      0
#> 369  -1.790369805      0
#> 370  -1.273977288      0
#> 371   0.793990872      0
#> 372   0.501832182      0
#> 373   1.321381043      0
#> 374  -0.018621808      0
#> 375  -1.119520773      0
#> 376  -0.421541106      0
#> 377  -1.251450693      0
#> 378  -0.443031019      0
#> 379  -0.334418125      0
#> 380   0.613477973      0
#> 381  -0.291269719      0
#> 382   0.139444655      0
#> 383  -0.518502332      0
#> 384   0.437108059      0
#> 385   2.325341452      0
#> 386   0.604919112      0
#> 387  -0.213513262      0
#> 388  -0.301847273      0
#> 389  -0.172662396      0
#> 390   0.931695346      0
#> 391  -0.271077560      0
#> 392  -1.459025941      0
#> 393  -0.175749220      0
#> 394  -1.232638831      0
#> 395   1.325132846      0
#> 396  -0.620407938      0
#> 397   0.140426766      0
#> 398   0.159484382      0
#> 399  -0.820718769      0
#> 400   1.653953159      0
#> 401  -0.010440059      0
#> 402  -0.776455844      0
#> 403  -0.876187344      0
#> 404  -0.356416456      0
#> 405  -0.878267793      0
#> 406  -0.612488499      0
#> 407   1.071119626      0
#> 408  -1.331417218      0
#> 409  -0.509196095      0
#> 410  -1.078116792      0
#> 411  -1.837407005      0
#> 412   1.782614617      0
#> 413   0.405868882      0
#> 414  -1.377573032      0
#> 415   0.755675661      0
#> 416   1.037290127      0
#> 417  -0.906661566      0
#> 418   0.099015898      0
#> 419  -0.780516665      0
#> 420  -0.037481319      0
#> 421  -2.125534319      0
#> 422   0.244155440      0
#> 423  -0.790655788      0
#> 424  -0.173957658      0
#> 425  -0.781202201      0
#> 426  -0.937897194      0
#> 427   0.015201561      0
#> 428   0.485834254      0
#> 429   0.659039519      0
#> 430  -2.242781281      0
#> 431  -0.140649980      0
#> 432  -0.847603839      0
#> 433  -0.301266927      0
#> 434  -0.455826634      0
#> 435   1.450036170      0
#> 436  -0.038130704      0
#> 437   1.470246314      0
#> 438   1.184731024      0
#> 439  -0.426793737      0
#> 440  -0.313851903      0
#> 441  -1.573228118      0
#> 442  -0.805901964      0
#> 443  -0.232329073      0
#> 444   0.587197539      0
#> 445  -0.571771550      0
#> 446   0.195337647      0
#> 447   0.343245254      0
#> 448  -1.080796803      0
#> 449  -0.877248968      0
#> 450  -0.972886909      0
#> 451   0.096772809      0
#> 452  -0.291308834      0
#> 453  -2.021667704      0
#> 454  -0.095404452      0
#> 455  -0.401908482      0
#> 456   0.607993509      0
#> 457   0.487883063      0
#> 458   1.007838527      0
#> 459  -0.328425067      0
#> 460  -0.486761947      0
#> 461  -2.013919632      0
#> 462  -0.251853414      0
#> 463   1.472584738      0
#> 464   1.290374809      0
#> 465  -0.663876660      0
#> 466   2.686221166      0
#> 467  -0.877130981      0
#> 468  -0.166720406      0
#> 469   0.110493627      0
#> 470  -3.419543157      0
#> 471   0.769391503      0
#> 472   0.255655459      0
#> 473   0.330677209      0
#> 474  -0.222250472      0
#> 475   0.554451344      0
#> 476  -0.592886425      0
#> 477   1.177901291      0
#> 478   0.653872386      0
#> 479   0.780423031      0
#> 480   0.449625816      0
#> 481  -0.474194924      0
#> 482  -0.048144642      0
#> 483  -0.679014300      0
#> 484   1.248243322      0
#> 485  -1.010808707      0
#> 486   0.182848590      0
#> 487  -0.420924962      0
#> 488   0.568717822      0
#> 489  -0.144522504      0
#> 490   0.650157957      0
#> 491   0.108683843      0
#> 492   1.933318120      0
#> 493   1.246012323      0
#> 494   1.309582577      0
#> 495   1.257580114      0
#> 496   1.275996260      0
#> 497  -0.331250550      0
#> 498  -0.025584508      0
#> 499   0.396490261      0
#> 500   1.368281760      0
#> 501   0.324541267      0
#> 502   1.174296915      0
#> 503  -0.739450930      0
#> 504  -0.442605964      0
#> 505   0.814062534      0
#> 506  -0.402995360      0
#> 507  -0.670365964      0
#> 508   0.166146153      0
#> 509  -1.383481181      0
#> 510  -1.019758530      0
#> 511  -0.236316345      0
#> 512  -0.478989467      0
#> 513   0.449320803      0
#> 514  -1.387732346      0
#> 515  -0.933011045      0
#> 516  -0.169710966      0
#> 517   2.134060567      0
#> 518  -0.568468601      0
#> 519  -0.117904244      0
#> 520   0.978601856      0
#> 521   0.776955114      0
#> 522   1.450521465      0
#> 523   0.316267157      0
#> 524  -0.935864332      0
#> 525   0.044162474      0
#> 526  -0.329578393      0
#> 527  -1.896634295      0
#> 528   0.190424436      0
#> 529  -0.560644441      0
#> 530  -0.766181381      0
#> 531  -1.123578742      0
#> 532  -0.909867879      0
#> 533  -0.203058078      0
#> 534   0.528214421      0
#> 535  -0.499078692      0
#> 536  -0.987085134      0
#> 537  -2.815690606      0
#> 538   0.746421718      0
#> 539  -0.123859674      0
#> 540   0.068479832      0
#> 541  -1.414044338      0
#> 542  -0.733103882      0
#> 543   0.098986170      0
#> 544  -0.452251481      0
#> 545  -1.351785334      0
#> 546   0.107292770      0
#> 547  -0.305594592      0
#> 548  -0.018117349      0
#> 549   0.985265534      0
#> 550  -0.296900383      0
#> 551  -0.211966998      0
#> 552   0.813157314      0
#> 553   1.265567975      0
#> 554  -0.463177737      0
#> 555   0.287801129      0
#> 556   1.522578782      0
#> 557  -0.501386941      0
#> 558   0.322717391      0
#> 559   0.475057030      0
#> 560  -0.543569669      0
#> 561   0.275417608      0
#> 562  -0.360763053      0
#> 563  -0.352725138      0
#> 564   2.038673063      0
#> 565   1.791271457      0
#> 566  -1.162717923      0
#> 567   0.252067937      0
#> 568   0.727030497      0
#> 569   0.206234092      0
#> 570   1.343170986      0
#> 571   0.590877161      0
#> 572  -0.811812877      0
#> 573  -0.169723166      0
#> 574  -0.346779982      0
#> 575   0.732767762      0
#> 576   0.438224089      0
#> 577   1.029095677      0
#> 578  -0.487522273      0
#> 579  -0.103713482      0
#> 580  -0.165661220      0
#> 581  -0.609187354      0
#> 582  -0.904502441      0
#> 583   0.439037618      0
#> 584   1.364270567      0
#> 585   1.246700871      0
#> 586   0.838373918      0
#> 587  -1.742676901      0
#> 588  -0.818146769      0
#> 589  -1.424822560      0
#> 590   0.504136668      0
#> 591   1.336196905      0
#> 592  -0.699256178      0
#> 593  -1.839394294      0
#> 594   0.868715475      0
#> 595   1.182287147      0
#> 596  -0.272049014      0
#> 597   0.129622659      0
#> 598  -0.003650256      0
#> 599   0.246490629      0
#> 600  -0.751500140      0
#> 601   0.452269922      0
#> 602  -0.124663867      0
#> 603  -2.340603945      0
#> 604   0.532139480      0
#> 605   1.371511599      0
#> 606   0.027157451      0
#> 607  -0.655563312      0
#> 608  -1.260676381      0
#> 609  -1.358394979      0
#> 610  -0.103144164      0
#> 611  -0.342676066      0
#> 612  -1.726520261      0
#> 613   1.063120926      0
#> 614  -1.599548960      0
#> 615   1.140993987      0
#> 616   2.317769738      0
#> 617   0.677295023      0
#> 618   1.538175829      0
#> 619  -0.529029327      0
#> 620  -0.496384745      0
#> 621   0.119794627      0
#> 622   0.081696434      0
#> 623  -0.061457134      0
#> 624   1.222178399      0
#> 625  -1.444505682      0
#> 626   1.346058942      0
#> 627   0.388137770      0
#> 628  -1.502499004      0
#> 629   0.620140800      0
#> 630  -1.253168484      0
#> 631   0.442175594      0
#> 632  -1.166073414      0
#> 633  -0.380917834      0
#> 634  -0.392295496      0
#> 635   0.772221106      0
#> 636  -0.551586736      0
#> 637   0.129671979      0
#> 638   0.990745010      0
#> 639   0.498702709      0
#> 640   0.973143417      0
#> 641  -1.124036917      0
#> 642   1.999457875      0
#> 643   0.881365014      0
#> 644   1.061261979      0
#> 645   2.306838078      0
#> 646   1.091439647      0
#> 647  -0.016519553      0
#> 648   0.904334029      0
#> 649   2.708676157      0
#> 650   0.661223874      0
#> 651   1.453630614      0
#> 652   0.140850386      0
#> 653  -0.684571347      0
#> 654   1.984146606      0
#> 655   0.317384616      0
#> 656   0.447088362      0
#> 657  -0.547082198      0
#> 658  -0.266197676      0
#> 659   1.124264751      0
#> 660  -0.284682278      0
#> 661  -0.283972620      0
#> 662  -0.051687020      0
#> 663  -0.895018455      0
#> 664  -0.143680400      0
#> 665   0.598189016      0
#> 666  -1.381391585      0
#> 667  -0.296370101      0
#> 668  -0.155401369      0
#> 669   0.627473736      0
#> 670  -0.249508632      0
#> 671   0.801332275      0
#> 672  -0.367884414      0
#> 673  -0.537075070      0
#> 674   0.673927803      0
#> 675   3.366311180      0
#> 676  -0.141141552      0
#> 677   1.329545450      0
#> 678  -1.142303292      0
#> 679  -0.024709932      0
#> 680   0.640860319      0
#> 681  -0.224054646      0
#> 682  -0.366960123      0
#> 683  -0.597527870      0
#> 684   0.304392634      0
#> 685   1.842530841      0
#> 686  -0.287954172      0
#> 687   0.525029917      0
#> 688   0.898334731      0
#> 689  -0.274343886      0
#> 690  -0.865624044      0
#> 691  -0.427221662      0
#> 692   0.526323034      0
#> 693   0.386288853      0
#> 694   1.500454997      0
#> 695   0.115463013      0
#> 696   0.094975587      0
#> 697   0.968747527      0
#> 698   0.666432058      0
#> 699   1.394905649      0
#> 700   0.120781401      0
#> 701  -0.189454550      0
#> 702  -0.597727033      0
#> 703  -0.471408042      0
#> 704   0.286660317      0
#> 705  -1.518243461      0
#> 706   1.044085783      0
#> 707  -1.157770574      0
#> 708   0.562874703      0
#> 709   0.948837553      0
#> 710   1.805990044      0
#> 711   0.799240243      0
#> 712  -0.736503812      0
#> 713  -0.754195523      0
#> 714   0.970417354      0
#> 715   0.053177635      0
#> 716   0.430392766      0
#> 717  -0.984167092      0
#> 718  -0.309679589      0
#> 719   1.486357305      0
#> 720  -0.849724334      0
#> 721  -0.230340854      0
#> 722  -0.764892958      0
#> 723  -0.113889367      0
#> 724   0.280779078      0
#> 725  -1.101629323      0
#> 726   0.398618808      0
#> 727   0.371829366      0
#> 728   0.128339877      0
#> 729   1.342806429      0
#> 730   0.931245401      0
#> 731   0.264542101      0
#> 732   1.388815145      0
#> 733   1.642743692      0
#> 734   0.751020997      0
#> 735  -0.583092452      0
#> 736  -1.232187413      0
#> 737   1.330330424      0
#> 738   1.040121777      0
#> 739  -1.496739230      0
#> 740   1.004974787      0
#> 741   0.439713544      0
#> 742   0.479778667      0
#> 743  -1.139043892      0
#> 744   1.141642635      0
#> 745  -1.367696787      0
#> 746   0.951361928      0
#> 747  -0.935035900      0
#> 748   0.293906331      0
#> 749   1.092400325      0
#> 750  -0.646855666      0
#> 751   0.214888630      0
#> 752  -0.309206232      0
#> 753   2.291100826      0
#> 754   0.880170511      0
#> 755   1.851171466      0
#> 756   0.674612948      0
#> 757   0.784636009      0
#> 758  -0.898635015      0
#> 759   0.029402883      0
#> 760   1.961693550      0
#> 761  -0.234831435      0
#> 762  -1.029491049      0
#> 763  -0.989789941      0
#> 764  -0.025806570      0
#> 765   0.657045314      0
#> 766   0.400987275      0
#> 767   0.002220488      0
#> 768   1.823831405      0
#> 769  -0.827299250      0
#> 770   0.399057033      0
#> 771  -0.350171226      0
#> 772  -0.016243157      0
#> 773  -0.275514000      0
#> 774  -1.572135945      0
#> 775   1.053647080      0
#> 776  -0.082037812      0
#> 777  -0.601337660      0
#> 778   0.890315129      0
#> 779   0.690228473      0
#> 780  -0.472561929      0
#> 781   1.456138357      0
#> 782  -0.565620098      0
#> 783   2.432524263      0
#> 784  -0.456619479      0
#> 785   0.560410015      0
#> 786   0.712256554      0
#> 787  -0.481463285      0
#> 788  -0.829629935      0
#> 789  -0.107835403      0
#> 790  -1.630913999      0
#> 791  -0.623863100      0
#> 792  -0.369318577      0
#> 793  -0.997354630      0
#> 794   0.984908833      0
#> 795   0.310621973      0
#> 796  -1.002863312      0
#> 797   0.281083397      0
#> 798   0.818913315      0
#> 799   0.073690221      0
#> 800   0.546721111      0
#> 801  -0.995038741      0
#> 802   1.349235272      0
#> 803   0.843095805      0
#> 804  -0.117308076      0
#> 805  -0.953256650      0
#> 806   0.394696440      0
#> 807   0.255755928      0
#> 808  -2.695464955      0
#> 809   1.157917172      0
#> 810   0.055389585      0
#> 811   0.153755776      0
#> 812   0.028381624      0
#> 813  -0.420273044      0
#> 814   0.132016727      0
#> 815  -1.460417925      0
#> 816  -0.219511712      0
#> 817   1.985804316      0
#> 818  -0.854663892      0
#> 819   0.861196781      0
#> 820  -1.015563989      0
#> 821   1.193134417      0
#> 822  -0.498693027      0
#> 823  -0.766071241      0
#> 824  -1.602124279      0
#> 825  -1.271735629      0
#> 826   2.001011498      0
#> 827   0.253665769      0
#> 828   0.445916498      0
#> 829  -0.427560462      0
#> 830   0.841249992      0
#> 831   1.167090788      0
#> 832   1.329379111      0
#> 833   0.077862936      0
#> 834  -0.364242676      0
#> 835  -0.168634074      0
#> 836  -1.790858112      0
#> 837  -1.001021871      0
#> 838  -1.599136813      0
#> 839  -0.056132615      0
#> 840  -0.482717374      0
#> 841   1.074365869      0
#> 842  -1.024909940      0
#> 843   0.651147972      0
#> 844  -0.419001277      0
#> 845  -0.189982427      0
#> 846   0.760664168      0
#> 847   0.203306683      0
#> 848   0.884067568      0
#> 849   0.325219669      0
#> 850   1.270571858      0
#> 851   0.756170379      0
#> 852  -1.701030061      0
#> 853   0.814858758      0
#> 854   1.028750821      0
#> 855  -0.478806990      0
#> 856   1.650254513      0
#> 857   0.368281371      0
#> 858  -1.710449272      0
#> 859  -1.572081842      0
#> 860  -0.811212642      0
#> 861   0.872472197      0
#> 862   0.322238220      0
#> 863   1.309450142      0
#> 864  -0.483676635      0
#> 865   0.626045112      0
#> 866  -2.101143910      0
#> 867  -2.500741101      0
#> 868   0.233149698      0
#> 869   0.061924705      0
#> 870   0.167654662      0
#> 871   0.834661863      0
#> 872   0.982217393      0
#> 873  -1.514342641      0
#> 874   0.156529553      0
#> 875   0.024344803      0
#> 876  -0.492307326      0
#> 877  -2.394054321      0
#> 878  -1.205084070      0
#> 879  -0.199517072      0
#> 880  -0.848679150      0
#> 881   0.618467488      0
#> 882  -0.084482571      0
#> 883  -0.303377576      0
#> 884  -0.920359470      0
#> 885  -0.492373328      0
#> 886   0.324538473      0
#> 887   0.762984073      0
#> 888  -0.003277563      1
#> 889   0.016297899      1
#> 890  -0.478198856      1
#> 891   1.887830932      1
#> 892  -0.371810258      1
#> 893   1.141595871      1
#> 894   1.600849517      1
#> 895   0.229439671      1
#> 896   0.867647815      1
#> 897  -1.660637695      1
#> 898   1.132224019      1
#> 899  -0.006910085      1
#> 900   1.601125421      1
#> 901   0.645620609      1
#> 902   2.246377805      1
#> 903   0.108273348      1
#> 904   0.483102282      1
#> 905   1.143039831      1
#> 906   0.177816620      1
#> 907   0.720915232      1
#> 908  -0.785737975      1
#> 909   1.875330842      1
#> 910   0.492748811      1
#> 911   0.689615161      1
#> 912   0.809024885      1
#> 913   2.783011594      1
#> 914   1.198105375      1
#> 915   0.871955472      1
#> 916   0.630637559      1
#> 917   0.009134930      1
#> 918   0.327335823      1
#> 919   1.230242414      1
#> 920   1.137494598      1
#> 921  -1.135265600      1
#> 922   0.731645872      1
#> 923   1.025551761      1
#> 924   0.611102584      1
#> 925  -0.769361111      1
#> 926   0.698690053      1
#> 927   1.885021147      1
#> 928   0.468680242      1
#> 929   0.881599555      1
#> 930  -0.085455031      1
#> 931   0.065583955      1
#> 932   1.094545398      1
#> 933   0.840671078      1
#> 934   0.935832881      1
#> 935   0.894776599      1
#> 936   0.126365498      1
#> 937   0.708773267      1
#> 938   1.077773345      1
#> 939   0.884328578      1
#> 940   0.947944784      1
#> 941  -0.919981284      1
#> 942   2.015017909      1
#> 943   1.271118201      1
#> 944  -1.230097672      1
#> 945   3.507746382      1
#> 946   1.023220024      1
#> 947   0.879422524      1
#> 948   3.525981606      1
#> 949   2.106723910      1
#> 950   2.223137077      1
#> 951  -0.649635046      1
#> 952   1.631596614      1
#> 953   0.308462972      1
#> 954   0.647707392      1
#> 955  -0.082094678      1
#> 956  -0.125061930      1
#> 957   2.474190353      1
#> 958   1.188888649      1
#> 959   0.455671022      1
#> 960   1.760199523      1
#> 961   0.504765245      1
#> 962   0.830400261      1
#> 963   1.569352859      1
#> 964   0.351686754      1
#> 965   1.653493683      1
#> 966   0.946130349      1
#> 967   1.490128195      1
#> 968   0.085087262      1
#> 969   0.224706192      1
#> 970   0.861560417      1
#> 971   1.776852112      1
#> 972   2.461333437      1
#> 973   2.858943194      1
#> 974   1.980205766      1
#> 975  -0.551800639      1
#> 976   0.298723237      1
#> 977   1.614517068      1
#> 978   0.694467847      1
#> 979  -0.082542741      1
#> 980   2.028614085      1
#> 981  -0.241571989      1
#> 982  -0.120641805      1
#> 983   0.667248287      1
#> 984  -0.072940307      1
#> 985   0.750663653      1
#> 986   1.298775335      1
#> 987   0.346128498      1
#> 988   0.636912530      1
#> 989   1.080706741      1
#> 990   1.709113215      1
#> 991   1.326097471      1
#> 992   2.645294771      1
#> 993  -0.676728150      1
#> 994   0.105304187      1
#> 995   0.260120072      1
#> 996   2.060609377      1
#> 997   0.954987491      1
#> 998  -1.334226287      1
#> 999   0.531303203      1
#> 1000 -0.018161748      1