Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.138334286      0
#> 2     1.844675092      0
#> 3     0.603894629      0
#> 4     0.550912091      0
#> 5     0.068014082      0
#> 6     1.025152620      0
#> 7    -1.264250064      0
#> 8    -0.672998243      0
#> 9     0.971625647      0
#> 10   -0.456346896      0
#> 11   -2.163012726      0
#> 12   -0.323324957      0
#> 13   -0.238335683      0
#> 14   -0.464382217      0
#> 15   -2.393215784      0
#> 16    0.820965366      0
#> 17   -1.550637983      0
#> 18   -0.362269850      0
#> 19   -1.670281679      0
#> 20   -0.428116466      0
#> 21    0.188900183      0
#> 22    0.036773253      0
#> 23    0.062995425      0
#> 24    0.942995851      0
#> 25   -1.039326653      0
#> 26    0.638712332      0
#> 27   -0.747751758      0
#> 28   -1.492925745      0
#> 29    0.791084011      0
#> 30   -1.182540827      0
#> 31   -1.473452147      0
#> 32    1.834079192      0
#> 33   -0.012382801      0
#> 34    0.018825262      0
#> 35   -2.323181951      0
#> 36   -0.103120408      0
#> 37   -1.514833610      0
#> 38    0.634687615      0
#> 39    0.473771585      0
#> 40   -0.458743701      0
#> 41   -0.165065090      0
#> 42    0.237374113      0
#> 43    0.705850554      0
#> 44    2.073767682      0
#> 45   -0.922312017      0
#> 46    0.478889212      0
#> 47   -0.781054636      0
#> 48   -0.150379530      0
#> 49    0.198503973      0
#> 50    1.912222140      0
#> 51    0.495237027      0
#> 52   -0.675740656      0
#> 53    0.214414730      0
#> 54    1.192825927      0
#> 55    1.718005058      0
#> 56    0.433492917      0
#> 57   -0.885710774      0
#> 58   -0.418271827      0
#> 59   -1.003361179      0
#> 60    0.015977853      0
#> 61   -0.257090399      0
#> 62   -0.993356766      0
#> 63    0.427177084      0
#> 64    1.001803674      0
#> 65   -1.681608774      0
#> 66    2.100194596      0
#> 67    0.028125914      0
#> 68   -0.126244004      0
#> 69    1.277959372      0
#> 70   -0.002011657      0
#> 71   -0.123645039      0
#> 72    1.257313496      0
#> 73   -0.413238536      0
#> 74    1.211642186      0
#> 75   -0.542176212      0
#> 76    0.544632429      0
#> 77   -1.201486890      0
#> 78   -0.898719436      0
#> 79   -0.070416419      0
#> 80   -1.477188418      0
#> 81   -0.287600052      0
#> 82    0.301994232      0
#> 83    2.039904469      0
#> 84   -1.980571224      0
#> 85   -1.174907875      0
#> 86   -0.879586670      0
#> 87   -0.704439650      0
#> 88    0.903550770      0
#> 89    0.077894377      0
#> 90    0.285739131      0
#> 91    1.039463931      0
#> 92    0.414760491      0
#> 93    1.546176232      0
#> 94    0.440679543      0
#> 95    1.961118651      0
#> 96    0.485884621      0
#> 97    0.190854979      0
#> 98   -0.856041011      0
#> 99    0.776090531      0
#> 100   0.714430731      0
#> 101  -0.272024226      0
#> 102  -1.014467206      0
#> 103   0.311154653      0
#> 104   0.298570332      0
#> 105  -0.420724835      0
#> 106   0.324786570      0
#> 107  -1.401762487      0
#> 108   0.921795274      0
#> 109  -0.730988483      0
#> 110  -0.519550509      0
#> 111  -0.337392134      0
#> 112  -0.313733000      0
#> 113  -1.600888504      0
#> 114  -0.138648052      0
#> 115  -0.598010952      0
#> 116   0.449296096      0
#> 117   0.179254534      0
#> 118  -0.535540641      0
#> 119   0.724804176      0
#> 120  -1.173880597      0
#> 121  -0.867263783      0
#> 122  -0.473028054      0
#> 123  -0.154343877      0
#> 124  -0.165369479      0
#> 125  -0.354938145      0
#> 126   1.961861035      0
#> 127  -0.223834564      0
#> 128  -0.692771655      0
#> 129   0.464191233      0
#> 130   0.076595374      0
#> 131  -0.077780941      0
#> 132  -0.870994437      0
#> 133   1.352863176      0
#> 134   0.125367408      0
#> 135  -2.249484640      0
#> 136   1.753678224      0
#> 137   0.490100171      0
#> 138   2.464462200      0
#> 139   0.852836739      0
#> 140   2.116575040      0
#> 141  -0.170150491      0
#> 142   0.674164242      0
#> 143  -0.217780933      0
#> 144  -1.225093460      0
#> 145   0.888568917      0
#> 146   2.057394785      0
#> 147  -2.342980994      0
#> 148  -0.410253231      0
#> 149  -1.313635572      0
#> 150  -0.134323719      0
#> 151  -0.452055412      0
#> 152  -1.155132467      0
#> 153   0.165341565      0
#> 154   2.880916894      0
#> 155   0.402191977      0
#> 156  -0.113201952      0
#> 157  -0.520695469      0
#> 158  -0.208463844      0
#> 159   0.916226681      0
#> 160  -0.462150313      0
#> 161  -0.232687576      0
#> 162   1.588149443      0
#> 163   0.294745643      0
#> 164   1.760476738      0
#> 165   1.858915484      0
#> 166   0.380390155      0
#> 167   0.785284343      0
#> 168  -0.410209921      0
#> 169  -0.441261165      0
#> 170   1.673729363      0
#> 171   2.178616808      0
#> 172  -0.817377976      0
#> 173   1.281451313      0
#> 174   1.911910139      0
#> 175   0.919277045      0
#> 176   0.615244576      0
#> 177  -0.658027662      0
#> 178   1.376403672      0
#> 179  -1.286720211      0
#> 180   0.070820439      0
#> 181   0.572759509      0
#> 182  -0.117049280      0
#> 183  -1.276507052      0
#> 184   1.203520698      0
#> 185   0.010452831      0
#> 186  -1.306931843      0
#> 187   0.522575024      0
#> 188  -0.915367735      0
#> 189  -1.246620345      0
#> 190   0.172038845      0
#> 191   0.781654351      0
#> 192  -0.798774566      0
#> 193   0.729522807      0
#> 194  -0.302448843      0
#> 195   0.859835574      0
#> 196   0.288671341      0
#> 197   0.232073273      0
#> 198  -1.092535408      0
#> 199  -0.006347926      0
#> 200   0.626642879      0
#> 201   0.254051179      0
#> 202  -0.226371622      0
#> 203   1.344699724      0
#> 204  -3.319301465      0
#> 205   0.336434399      0
#> 206   1.446163546      0
#> 207  -1.536731553      0
#> 208  -0.219447148      0
#> 209  -1.004469855      0
#> 210   1.785129425      0
#> 211   0.966740938      0
#> 212   1.635313833      0
#> 213   0.020868247      0
#> 214   1.535656107      0
#> 215   1.299381047      0
#> 216   0.299277422      0
#> 217  -0.011901498      0
#> 218   0.226583742      0
#> 219   0.099648936      0
#> 220  -0.787098292      0
#> 221   1.921777537      0
#> 222   0.586007345      0
#> 223  -2.165852927      0
#> 224   0.202037192      0
#> 225   0.547152768      0
#> 226   0.789035610      0
#> 227   1.088108066      0
#> 228  -0.936907442      0
#> 229   0.466985905      0
#> 230  -0.356246496      0
#> 231   1.393711611      0
#> 232  -0.063674034      0
#> 233   0.969534525      0
#> 234   1.072621710      0
#> 235  -1.029002785      0
#> 236  -1.374043817      0
#> 237  -1.105456296      0
#> 238   0.014428824      0
#> 239  -0.448040871      0
#> 240   0.959654863      0
#> 241   0.201254539      0
#> 242   1.207180934      0
#> 243  -1.411581304      0
#> 244  -0.090483100      0
#> 245   1.469281069      0
#> 246  -1.080992583      0
#> 247  -1.018198412      0
#> 248  -0.109456653      0
#> 249   0.448388306      0
#> 250  -0.659726798      0
#> 251  -0.987158958      0
#> 252   0.732992165      0
#> 253   0.895197902      0
#> 254  -0.738052122      0
#> 255   1.236177663      0
#> 256   0.296144829      0
#> 257  -0.423337607      0
#> 258   0.240937529      0
#> 259   0.191513317      0
#> 260  -1.394500690      0
#> 261   1.555722570      0
#> 262  -0.131501531      0
#> 263  -4.085026321      0
#> 264  -0.661140023      0
#> 265   2.150817733      0
#> 266  -1.600028711      0
#> 267   1.145376383      0
#> 268  -0.497158973      0
#> 269  -0.715396641      0
#> 270  -0.671344524      0
#> 271  -1.121608357      0
#> 272  -0.151174491      0
#> 273   1.846465852      0
#> 274  -1.499916000      0
#> 275   0.742697508      0
#> 276  -0.392177802      0
#> 277  -1.224439261      0
#> 278   1.417164311      0
#> 279  -0.451007403      0
#> 280   0.946008666      0
#> 281  -0.053249949      0
#> 282  -0.049481265      0
#> 283  -0.476174273      0
#> 284   0.776869014      0
#> 285   0.608016769      0
#> 286   0.119821949      0
#> 287  -0.074485054      0
#> 288   0.219966563      0
#> 289  -0.078011364      0
#> 290  -0.712824175      0
#> 291  -0.799684964      0
#> 292  -0.624326449      0
#> 293  -0.654802815      0
#> 294   1.509991192      0
#> 295  -0.727886392      0
#> 296  -0.132391216      0
#> 297  -0.138710612      0
#> 298   0.725682637      0
#> 299  -0.989058141      0
#> 300  -0.950877164      0
#> 301  -0.647690042      0
#> 302  -1.026272303      0
#> 303   0.134408345      0
#> 304  -0.030127205      0
#> 305  -1.593863930      0
#> 306   0.409329302      0
#> 307  -0.630356439      0
#> 308  -0.798013942      0
#> 309  -0.020950221      0
#> 310   0.415887043      0
#> 311   1.646817737      0
#> 312  -0.918290417      0
#> 313  -1.422468366      0
#> 314  -1.005204113      0
#> 315  -1.376927162      0
#> 316  -0.779327505      0
#> 317   0.611832069      0
#> 318  -0.461072163      0
#> 319   0.656147468      0
#> 320   1.476649075      0
#> 321   1.481469713      0
#> 322   1.863630910      0
#> 323  -0.031904716      0
#> 324   0.633264475      0
#> 325   0.993114792      0
#> 326  -0.700439568      0
#> 327   0.677286435      0
#> 328  -0.801907696      0
#> 329   1.051386781      0
#> 330  -0.579837424      0
#> 331   1.434767964      0
#> 332  -1.015339462      0
#> 333  -0.160788656      0
#> 334  -1.150852930      0
#> 335  -1.278825792      0
#> 336   1.333160111      0
#> 337   0.031220498      0
#> 338   1.162318727      0
#> 339   1.126723012      0
#> 340   0.243423441      0
#> 341  -0.056507627      0
#> 342  -0.423826229      0
#> 343  -0.050250756      0
#> 344   0.068088706      0
#> 345  -1.038226162      0
#> 346  -0.921250266      0
#> 347   1.079039886      0
#> 348  -0.031777715      0
#> 349  -2.180666611      0
#> 350   0.225524093      0
#> 351  -1.358303180      0
#> 352  -0.423783221      0
#> 353  -0.797478141      0
#> 354   1.334858127      0
#> 355   0.612285865      0
#> 356   0.037755758      0
#> 357  -0.348643132      0
#> 358   1.759730478      0
#> 359   0.230479942      0
#> 360  -0.009352563      0
#> 361   0.257956509      0
#> 362  -0.491310311      0
#> 363  -0.840616484      0
#> 364  -0.918812509      0
#> 365   1.151707871      0
#> 366  -0.098458104      0
#> 367  -0.888393412      0
#> 368   1.577733103      0
#> 369   2.220738014      0
#> 370   1.295621517      0
#> 371  -1.943195272      0
#> 372   0.394110341      0
#> 373  -0.663190495      0
#> 374   0.788015322      0
#> 375  -1.646424486      0
#> 376   0.395362074      0
#> 377   0.536724353      0
#> 378   0.858837180      0
#> 379  -1.362680179      0
#> 380  -0.224691592      0
#> 381  -0.970424625      0
#> 382   0.476581498      0
#> 383   0.510476039      0
#> 384   0.425500435      0
#> 385   0.042446790      0
#> 386  -0.073292349      0
#> 387   1.429152860      0
#> 388   0.094278903      0
#> 389  -0.015624091      0
#> 390  -0.335348289      0
#> 391  -0.605762722      0
#> 392   0.642601357      0
#> 393  -0.197502606      0
#> 394  -0.823501898      0
#> 395  -1.144588450      0
#> 396   1.112923993      0
#> 397   0.122528639      0
#> 398  -0.050751114      0
#> 399  -1.046752784      0
#> 400   0.040575510      0
#> 401   1.349381305      0
#> 402   1.116162002      0
#> 403  -0.054539179      0
#> 404  -0.714270969      0
#> 405   0.040806236      0
#> 406  -0.050709118      0
#> 407  -0.552712211      0
#> 408  -0.613721294      0
#> 409  -1.186557894      0
#> 410   0.909687597      0
#> 411   0.204341380      0
#> 412   0.012545480      0
#> 413  -1.866283310      0
#> 414  -1.667920833      0
#> 415   0.568019083      0
#> 416  -0.499787599      0
#> 417   1.490073195      0
#> 418   0.782120240      0
#> 419   0.518974714      0
#> 420   0.044012275      0
#> 421  -1.773474766      0
#> 422  -1.249801532      0
#> 423   0.483737718      0
#> 424   2.147445637      0
#> 425  -1.262548690      0
#> 426  -0.309926770      0
#> 427  -0.606686913      0
#> 428   1.457180988      0
#> 429   0.287316797      0
#> 430  -0.729706088      0
#> 431  -2.459559154      0
#> 432  -0.195739736      0
#> 433  -0.017871829      0
#> 434   0.449830535      0
#> 435  -1.070029634      0
#> 436  -0.791356597      0
#> 437  -0.968517848      0
#> 438   0.308107792      0
#> 439   0.327453406      0
#> 440   0.002783829      0
#> 441  -1.227812571      0
#> 442  -2.348418545      0
#> 443  -0.966937714      0
#> 444   1.805185712      0
#> 445  -1.471744555      0
#> 446   0.568109976      0
#> 447   0.006620076      0
#> 448   0.484843207      0
#> 449  -1.221842484      0
#> 450   1.398694343      0
#> 451   1.416559981      0
#> 452   0.072530017      0
#> 453  -0.671106195      0
#> 454   0.107268314      0
#> 455  -1.373590129      0
#> 456  -0.095588134      0
#> 457  -0.961624515      0
#> 458   0.392336311      0
#> 459   0.331735816      0
#> 460   0.020922345      0
#> 461   1.046341869      0
#> 462   0.672497879      0
#> 463   0.517231007      0
#> 464   0.740291849      0
#> 465  -0.693963014      0
#> 466  -2.305867713      0
#> 467  -0.055451825      0
#> 468  -2.671534602      0
#> 469  -0.213480726      0
#> 470  -0.136377749      0
#> 471   0.756938201      0
#> 472   0.532015417      0
#> 473   0.598290684      0
#> 474  -0.386141093      0
#> 475   0.305548830      0
#> 476   0.827772373      0
#> 477   0.966874040      0
#> 478   0.476513833      0
#> 479   0.119908365      0
#> 480   0.160931128      0
#> 481  -1.040729197      0
#> 482  -1.701332829      0
#> 483  -2.007343062      0
#> 484  -0.604862485      0
#> 485   1.492898862      0
#> 486  -0.556457145      0
#> 487  -0.797756719      0
#> 488  -2.372122862      0
#> 489   0.630705991      0
#> 490  -0.883109893      0
#> 491   2.001808110      0
#> 492  -0.728629504      0
#> 493   1.126976715      0
#> 494   0.493828920      0
#> 495   1.161728535      0
#> 496  -0.308415101      0
#> 497  -0.159101980      0
#> 498  -0.126517681      0
#> 499  -0.214031035      0
#> 500  -0.271556009      0
#> 501   2.142328216      0
#> 502   1.432483414      0
#> 503  -0.822805737      0
#> 504  -0.731540064      0
#> 505   1.498266409      0
#> 506  -0.798724797      0
#> 507  -0.141651305      0
#> 508  -0.801682630      0
#> 509   1.413926017      0
#> 510   0.529085317      0
#> 511  -0.942573031      0
#> 512  -0.346284413      0
#> 513  -0.626681320      0
#> 514  -0.459136393      0
#> 515  -2.611163193      0
#> 516   0.568588068      0
#> 517   0.845499386      0
#> 518  -1.649652519      0
#> 519  -0.087009064      0
#> 520  -0.363775334      0
#> 521  -1.691804685      0
#> 522  -0.997545634      0
#> 523   0.539978709      0
#> 524  -1.090530203      0
#> 525  -1.037200006      0
#> 526   1.752099867      0
#> 527  -0.388965461      0
#> 528  -0.235628729      0
#> 529  -1.421978770      0
#> 530  -1.134342889      0
#> 531  -0.252136233      0
#> 532  -2.140177809      0
#> 533  -1.373973453      0
#> 534   1.308707123      0
#> 535  -0.540653046      0
#> 536   0.947936873      0
#> 537  -0.101209714      0
#> 538   1.430105744      0
#> 539   1.001075042      0
#> 540  -2.411419672      0
#> 541   0.531020681      0
#> 542   1.760099632      0
#> 543  -0.366050164      0
#> 544  -0.105653975      0
#> 545  -1.669458634      0
#> 546  -1.084166928      0
#> 547  -0.239799304      0
#> 548   0.916702216      0
#> 549   0.132995901      0
#> 550   0.430771921      0
#> 551   2.119036576      0
#> 552   1.302086805      0
#> 553  -0.144924240      0
#> 554   0.066053263      0
#> 555  -0.367073239      0
#> 556  -0.144266930      0
#> 557  -0.003148326      0
#> 558   0.349905458      0
#> 559   1.860051586      0
#> 560  -0.133329375      0
#> 561  -0.381998429      0
#> 562  -0.360011634      0
#> 563  -0.496262955      0
#> 564   0.587716044      0
#> 565   0.865324617      0
#> 566   0.523366178      0
#> 567  -0.292810832      0
#> 568   0.672036696      0
#> 569  -1.562561836      0
#> 570  -0.020847655      0
#> 571  -0.490660932      0
#> 572  -0.303058499      0
#> 573   1.185449482      0
#> 574   1.229001401      0
#> 575  -0.332855199      0
#> 576  -1.292550054      0
#> 577   0.441000620      0
#> 578  -0.258621848      0
#> 579   0.215463451      0
#> 580  -1.441680269      0
#> 581  -0.134476617      0
#> 582  -1.051818591      0
#> 583  -0.297644610      0
#> 584  -1.311888068      0
#> 585   1.659736333      0
#> 586   1.536525607      0
#> 587  -1.313676258      0
#> 588   0.399769719      0
#> 589  -0.115700463      0
#> 590   0.635071910      0
#> 591  -0.137407011      0
#> 592  -0.428311515      0
#> 593  -0.823813725      0
#> 594   0.988266029      0
#> 595  -1.163363766      0
#> 596   0.543278976      0
#> 597  -0.306376702      0
#> 598   0.152055685      0
#> 599   0.527114334      0
#> 600   0.134392579      0
#> 601   1.710175775      0
#> 602  -2.391331562      0
#> 603  -0.188838678      0
#> 604   0.874087762      0
#> 605   0.591258021      0
#> 606   1.215154088      0
#> 607   0.500917816      0
#> 608  -1.474332227      0
#> 609   0.487421720      0
#> 610   0.299361876      0
#> 611  -0.034245962      0
#> 612   1.409845989      0
#> 613  -0.224088365      0
#> 614  -0.249789899      0
#> 615  -1.460083981      0
#> 616   0.594760080      0
#> 617  -0.407575515      0
#> 618  -0.590124192      0
#> 619  -0.702053136      0
#> 620  -0.367369770      0
#> 621   0.074077854      0
#> 622  -0.679192403      0
#> 623  -2.995281506      0
#> 624   1.128007573      0
#> 625   0.769273939      0
#> 626   0.284262879      0
#> 627   1.489723902      0
#> 628  -0.470408219      0
#> 629   1.456521260      0
#> 630  -1.351006650      0
#> 631   0.458840649      0
#> 632  -0.186436949      0
#> 633  -0.900733092      0
#> 634  -0.783671403      0
#> 635   1.172878017      0
#> 636  -0.061657688      0
#> 637  -1.140813005      0
#> 638   0.019288703      0
#> 639  -0.460820325      0
#> 640  -0.876154507      0
#> 641   1.707837801      0
#> 642  -0.713803756      0
#> 643  -0.367501371      0
#> 644   0.562440707      0
#> 645  -0.519204334      0
#> 646  -0.054659835      0
#> 647   0.252709923      0
#> 648   0.237544308      0
#> 649  -0.898172723      0
#> 650  -1.251227760      0
#> 651  -1.614091875      0
#> 652  -2.127422091      0
#> 653   0.692987072      0
#> 654  -2.860253467      0
#> 655  -1.304859942      0
#> 656  -0.455181283      0
#> 657  -0.098723303      0
#> 658  -0.748638425      0
#> 659  -0.593000643      0
#> 660  -0.572946733      0
#> 661   0.061557450      0
#> 662   0.905244072      0
#> 663  -0.630341389      0
#> 664  -0.563178975      0
#> 665   0.408521998      0
#> 666   0.421439271      0
#> 667  -0.090923997      0
#> 668   1.139557807      0
#> 669   0.064750606      0
#> 670   0.608331488      0
#> 671   0.573054725      0
#> 672   0.874090538      0
#> 673  -1.438387225      0
#> 674   1.553884214      0
#> 675  -1.624470456      0
#> 676  -0.953581769      0
#> 677  -0.051845263      0
#> 678   0.099728767      0
#> 679   1.108993192      0
#> 680   0.452410927      0
#> 681  -0.280388986      0
#> 682   1.025529490      0
#> 683  -0.712592803      0
#> 684   0.575192133      0
#> 685   0.476303158      0
#> 686  -1.313907433      0
#> 687   0.167234099      0
#> 688   0.696636413      0
#> 689   0.510378100      0
#> 690   0.251796705      0
#> 691   1.559342488      0
#> 692  -0.646917767      0
#> 693  -1.117490868      0
#> 694   0.192926299      0
#> 695  -0.384048721      0
#> 696   0.473442190      0
#> 697  -0.412337640      0
#> 698  -0.342217889      0
#> 699   0.733994957      0
#> 700   0.114254771      0
#> 701  -0.629836001      0
#> 702   0.019367586      0
#> 703   0.600245551      0
#> 704   0.524911023      0
#> 705   0.320819013      0
#> 706   0.506738944      0
#> 707  -0.815216348      0
#> 708  -1.085223457      0
#> 709  -0.533760001      0
#> 710   1.904827442      0
#> 711   0.351210312      0
#> 712  -1.206676992      0
#> 713   0.850905593      0
#> 714   0.207155291      0
#> 715  -0.883641427      0
#> 716  -0.101285282      0
#> 717  -0.320883290      0
#> 718  -1.151683498      0
#> 719  -0.462388464      0
#> 720  -1.318084607      0
#> 721  -0.048087187      0
#> 722  -1.178884228      0
#> 723   0.214038654      0
#> 724  -1.370364000      0
#> 725   0.210139511      0
#> 726   0.627628708      0
#> 727  -0.571570581      0
#> 728  -0.796716930      0
#> 729   0.790818858      0
#> 730   0.166584621      0
#> 731   1.249266730      0
#> 732   1.833868343      0
#> 733   0.093221751      0
#> 734  -0.943965031      0
#> 735  -0.585954757      0
#> 736   1.374343079      0
#> 737   0.335747219      0
#> 738  -0.340403280      0
#> 739  -0.415993569      0
#> 740  -1.005771863      0
#> 741  -0.647745197      0
#> 742   0.416828698      0
#> 743  -1.626756181      0
#> 744  -0.695524944      0
#> 745   0.107296232      0
#> 746  -0.886074644      0
#> 747   1.817229928      0
#> 748   0.044477954      0
#> 749  -0.463622329      0
#> 750   0.311935815      0
#> 751   0.513272585      0
#> 752   0.812946218      0
#> 753  -0.198852379      0
#> 754   0.765111220      0
#> 755  -0.624221136      0
#> 756  -1.416789271      0
#> 757  -0.244027446      0
#> 758  -1.180443315      0
#> 759  -0.554541713      0
#> 760   0.460938932      0
#> 761   0.722023006      0
#> 762  -1.019607614      0
#> 763  -0.357645492      0
#> 764  -0.170501417      0
#> 765   2.776306483      0
#> 766  -0.331959619      0
#> 767  -0.012254280      0
#> 768  -0.095286308      0
#> 769  -0.431600258      0
#> 770  -2.250845879      0
#> 771  -1.321151278      0
#> 772  -0.958007940      0
#> 773  -2.608248063      0
#> 774  -0.002196600      0
#> 775  -0.443128622      0
#> 776  -0.356944462      0
#> 777   0.119914716      0
#> 778  -0.425867346      0
#> 779  -1.226794368      0
#> 780   0.446263137      0
#> 781  -0.328039715      0
#> 782  -0.394495297      0
#> 783  -0.417085580      0
#> 784  -0.342145176      0
#> 785  -0.115834183      0
#> 786   0.480766286      0
#> 787  -0.438754858      0
#> 788  -1.204479498      0
#> 789   1.105441433      0
#> 790  -1.758300052      0
#> 791  -0.786247012      0
#> 792  -0.010217370      0
#> 793  -1.026138832      0
#> 794  -0.106047008      0
#> 795   1.588246760      0
#> 796   0.785052306      0
#> 797   1.385138894      0
#> 798  -0.240911641      0
#> 799  -1.523161482      0
#> 800   0.746961461      0
#> 801  -1.140047832      0
#> 802  -1.556227859      0
#> 803  -0.233849889      0
#> 804   2.383150106      0
#> 805   0.582040395      0
#> 806  -0.467268195      0
#> 807   0.902446673      0
#> 808   0.631739688      0
#> 809  -0.337671636      0
#> 810   0.400350596      0
#> 811  -1.036200136      0
#> 812  -1.689735248      0
#> 813  -0.311083191      0
#> 814  -0.553035017      0
#> 815   0.174679958      0
#> 816  -0.006903240      0
#> 817   0.397206164      0
#> 818   0.271987750      0
#> 819  -0.566859476      0
#> 820   0.796459149      0
#> 821  -0.496491780      0
#> 822  -0.619360114      0
#> 823   1.036021905      0
#> 824   1.071379892      0
#> 825  -0.588464784      0
#> 826   0.972441412      0
#> 827   0.315273872      0
#> 828  -1.343787908      0
#> 829   0.460114028      0
#> 830   0.469073783      0
#> 831  -0.021605321      0
#> 832   0.092714434      0
#> 833  -0.358280212      0
#> 834   0.125508213      0
#> 835  -0.674967175      0
#> 836   0.959929118      0
#> 837  -0.973554954      0
#> 838  -1.490757097      0
#> 839  -0.392273639      0
#> 840  -0.845781809      0
#> 841  -0.017291620      0
#> 842  -1.029171729      0
#> 843  -0.989261174      0
#> 844  -0.503253588      0
#> 845   0.977438125      0
#> 846   0.878453241      0
#> 847  -0.917737633      0
#> 848   0.408227834      0
#> 849  -0.309222495      0
#> 850   0.510950761      0
#> 851   0.263546711      0
#> 852  -1.687207925      0
#> 853  -1.256082874      0
#> 854   0.930480874      0
#> 855  -0.235505972      0
#> 856  -0.703737903      0
#> 857   0.255351536      0
#> 858  -0.832061538      0
#> 859   0.368044603      0
#> 860   0.594304681      0
#> 861   0.824411271      0
#> 862  -2.049819028      0
#> 863   0.739589543      0
#> 864   1.177954443      0
#> 865   0.183562973      0
#> 866   0.294011498      0
#> 867  -0.090967388      0
#> 868  -0.984877187      0
#> 869   0.206570023      0
#> 870  -2.481355477      0
#> 871  -0.007251530      0
#> 872  -1.633328127      0
#> 873  -1.046042123      0
#> 874  -0.135982371      0
#> 875   1.234477813      0
#> 876   0.707425018      0
#> 877  -1.608357264      0
#> 878   0.479100110      0
#> 879   0.272536934      0
#> 880   1.842308398      0
#> 881   0.849626640      0
#> 882  -1.084087221      0
#> 883   0.619386130      0
#> 884  -2.379371236      0
#> 885   1.733148564      0
#> 886   1.251976411      0
#> 887  -0.372576181      0
#> 888  -0.094926362      0
#> 889  -0.206629622      0
#> 890   2.076272645      1
#> 891  -0.051169518      1
#> 892   2.582879649      1
#> 893  -0.526050382      1
#> 894   0.680532256      1
#> 895   0.644530782      1
#> 896   1.464236401      1
#> 897   1.336377741      1
#> 898  -0.634987792      1
#> 899   1.645302839      1
#> 900   0.408673962      1
#> 901   1.135565506      1
#> 902   1.770612868      1
#> 903  -0.363514228      1
#> 904  -1.079964499      1
#> 905   1.198071303      1
#> 906   2.278508063      1
#> 907   2.223552584      1
#> 908  -0.368800873      1
#> 909  -0.104366686      1
#> 910   2.200720504      1
#> 911   1.331777771      1
#> 912   1.212190733      1
#> 913   1.041119546      1
#> 914  -0.269619967      1
#> 915   3.112295323      1
#> 916   1.759313213      1
#> 917  -0.483573475      1
#> 918   3.100581664      1
#> 919   0.732799903      1
#> 920   3.025079145      1
#> 921  -0.037717071      1
#> 922  -1.562131199      1
#> 923   1.551066640      1
#> 924   0.880604462      1
#> 925   1.371824234      1
#> 926   1.466531508      1
#> 927   0.569947759      1
#> 928   1.269257249      1
#> 929  -0.749777847      1
#> 930   0.802954759      1
#> 931   1.345821268      1
#> 932   0.369895849      1
#> 933   1.365557077      1
#> 934   2.061973643      1
#> 935   1.500699233      1
#> 936  -1.140756763      1
#> 937   0.664017450      1
#> 938   0.533545897      1
#> 939   1.297287726      1
#> 940   0.014045769      1
#> 941  -0.450878400      1
#> 942   1.860579841      1
#> 943   0.986344879      1
#> 944   1.862661301      1
#> 945  -0.696484066      1
#> 946   0.580591913      1
#> 947  -0.017733093      1
#> 948   1.597318781      1
#> 949   1.769534128      1
#> 950   1.849950966      1
#> 951   1.772909500      1
#> 952   1.319293330      1
#> 953   1.006425883      1
#> 954   0.401676447      1
#> 955   1.087157219      1
#> 956   0.420335419      1
#> 957   1.867019384      1
#> 958   1.811718655      1
#> 959   1.971522796      1
#> 960  -0.189711119      1
#> 961   2.652483144      1
#> 962   0.967528844      1
#> 963   0.903682866      1
#> 964   0.518143033      1
#> 965  -0.047514661      1
#> 966  -0.064987537      1
#> 967  -0.667727458      1
#> 968  -0.308378112      1
#> 969   0.294121594      1
#> 970   1.414051123      1
#> 971   2.668650166      1
#> 972   3.262611696      1
#> 973  -1.426247001      1
#> 974  -0.360389609      1
#> 975  -0.621024252      1
#> 976  -1.467187251      1
#> 977   1.524728258      1
#> 978   1.550153317      1
#> 979   1.396959898      1
#> 980  -1.668741390      1
#> 981   1.278212362      1
#> 982   0.649374824      1
#> 983   0.362681873      1
#> 984   0.994897475      1
#> 985   0.990240719      1
#> 986   2.355479018      1
#> 987  -0.190295443      1
#> 988   1.942133191      1
#> 989  -0.174620621      1
#> 990   1.980916583      1
#> 991   0.576057688      1
#> 992  -0.034147984      1
#> 993   0.382283198      1
#> 994   0.645747268      1
#> 995   2.950833856      1
#> 996   2.761849402      1
#> 997   1.432019317      1
#> 998   0.148704345      1
#> 999   1.037682685      1
#> 1000  0.853958542      1