Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     0.4820359648      0
#> 2     0.1436573203      0
#> 3     1.0147197466      0
#> 4    -0.9119882086      0
#> 5    -0.9310729423      0
#> 6     0.5852153697      0
#> 7    -0.6374452870      0
#> 8    -0.0432102728      0
#> 9     0.2654783238      0
#> 10   -0.3149973912      0
#> 11   -0.2330451632      0
#> 12   -0.0202072571      0
#> 13    0.4874102115      0
#> 14   -1.2731797169      0
#> 15    0.1581186819      0
#> 16   -0.4867407812      0
#> 17   -0.8501702800      0
#> 18   -0.0208169601      0
#> 19   -0.3756408594      0
#> 20    0.9889140317      0
#> 21   -0.1612038159      0
#> 22   -0.2964357552      0
#> 23    2.0559405321      0
#> 24   -0.6709960957      0
#> 25   -0.2315264387      0
#> 26    0.8601184657      0
#> 27   -0.1425045887      0
#> 28    0.8275422260      0
#> 29   -0.7301974076      0
#> 30    0.3882949775      0
#> 31   -1.0504063047      0
#> 32   -0.3421506261      0
#> 33   -0.7222398323      0
#> 34   -0.8248387818      0
#> 35   -0.3114175575      0
#> 36    0.4780827072      0
#> 37    0.1795500729      0
#> 38   -1.3125552019      0
#> 39    0.3720431819      0
#> 40    1.3322782211      0
#> 41   -0.1666308986      0
#> 42   -0.6276437675      0
#> 43    0.9698931359      0
#> 44    0.3066461125      0
#> 45    0.0141521906      0
#> 46   -0.5748932846      0
#> 47    0.0847709974      0
#> 48   -1.0512406683      0
#> 49   -1.2647907997      0
#> 50    1.2858744711      0
#> 51   -0.4772293744      0
#> 52    0.2759285192      0
#> 53   -0.7044713338      0
#> 54    2.0005250055      0
#> 55    1.9832000839      0
#> 56   -2.3192744998      0
#> 57   -0.8641788992      0
#> 58   -0.6405945689      0
#> 59    1.3400272972      0
#> 60    1.1475179579      0
#> 61    1.7227063688      0
#> 62   -1.7802678870      0
#> 63   -1.6331752608      0
#> 64    1.0744970629      0
#> 65    0.3603321829      0
#> 66   -2.0893854217      0
#> 67    1.1798817647      0
#> 68   -1.5512797392      0
#> 69    0.0990456157      0
#> 70    0.0468015149      0
#> 71    0.4101896574      0
#> 72   -0.4047593800      0
#> 73   -0.6761511807      0
#> 74    1.7011866810      0
#> 75   -1.2243784810      0
#> 76    0.3658393014      0
#> 77    1.9606411155      0
#> 78    0.7818547188      0
#> 79    1.7121687416      0
#> 80    1.3548435591      0
#> 81   -1.6144902758      0
#> 82   -1.6776558232      0
#> 83   -0.1201541343      0
#> 84    0.5493750354      0
#> 85   -1.3249514417      0
#> 86   -0.3771493913      0
#> 87    1.5246985625      0
#> 88    1.0603544049      0
#> 89    2.1251953604      0
#> 90    0.5858015436      0
#> 91    0.0323139762      0
#> 92    0.3463311112      0
#> 93   -1.0413247939      0
#> 94    0.1941157242      0
#> 95    1.8789249146      0
#> 96   -0.5560346836      0
#> 97    0.3550320946      0
#> 98    0.1918071880      0
#> 99   -0.3285626899      0
#> 100  -0.5221897635      0
#> 101  -0.5315742115      0
#> 102  -1.1656668261      0
#> 103  -0.6321075712      0
#> 104   0.9928071356      0
#> 105   1.1436757432      0
#> 106   1.7846276328      0
#> 107   0.0699010788      0
#> 108  -1.2432076666      0
#> 109  -0.6565308270      0
#> 110   0.6575251612      0
#> 111   0.0974591903      0
#> 112   1.5431732215      0
#> 113  -0.6954301973      0
#> 114   3.0928520319      0
#> 115   1.0090548941      0
#> 116  -1.5176757394      0
#> 117  -0.7089216107      0
#> 118   0.9534049935      0
#> 119   1.9964421931      0
#> 120   0.0262585320      0
#> 121   0.9061795101      0
#> 122  -0.7786610959      0
#> 123  -2.2980361219      0
#> 124  -0.9649604358      0
#> 125  -0.6377216727      0
#> 126  -0.1438944129      0
#> 127   1.5619921770      0
#> 128   0.6716102367      0
#> 129  -0.4108458247      0
#> 130  -0.0077165003      0
#> 131  -0.7338925020      0
#> 132   0.6067033109      0
#> 133   0.0306528428      0
#> 134   1.9786348822      0
#> 135  -0.2048046338      0
#> 136   0.2645065456      0
#> 137   0.2198597551      0
#> 138  -0.7783839378      0
#> 139   0.3764896518      0
#> 140   1.4455138395      0
#> 141  -1.3332728060      0
#> 142   1.0265010002      0
#> 143  -1.1889834503      0
#> 144   1.4613512743      0
#> 145  -1.6300060535      0
#> 146  -1.3971353480      0
#> 147  -0.1060928880      0
#> 148  -0.3370757002      0
#> 149   0.8985715907      0
#> 150   0.9654886673      0
#> 151  -0.4388256021      0
#> 152  -1.4993505804      0
#> 153   1.8357330759      0
#> 154  -0.0160501510      0
#> 155  -1.6447160080      0
#> 156   0.6383983444      0
#> 157  -1.5328701962      0
#> 158  -0.7776487426      0
#> 159  -0.2508968045      0
#> 160   2.2306255075      0
#> 161  -0.3380219598      0
#> 162  -0.9692882962      0
#> 163   0.8911043125      0
#> 164  -1.0837635517      0
#> 165   0.3636428743      0
#> 166  -0.6387795936      0
#> 167  -0.7593605929      0
#> 168  -1.0278902737      0
#> 169  -0.0376554006      0
#> 170   0.7248805576      0
#> 171   1.3419042072      0
#> 172   0.7215743415      0
#> 173  -0.7601547345      0
#> 174  -1.1260150078      0
#> 175  -1.5922235803      0
#> 176  -1.0717887463      0
#> 177   2.1037765693      0
#> 178  -0.8657422652      0
#> 179  -0.0164022802      0
#> 180  -0.5906337318      0
#> 181  -0.7455131309      0
#> 182  -1.4908527965      0
#> 183   1.4026605696      0
#> 184  -1.1332063276      0
#> 185   0.1402849939      0
#> 186   0.6205934677      0
#> 187   1.5346588274      0
#> 188   0.7102134774      0
#> 189  -0.6832732119      0
#> 190   0.2074755388      0
#> 191  -1.1929141063      0
#> 192   0.7375759445      0
#> 193  -0.1152308129      0
#> 194   0.1568020408      0
#> 195   0.3754002608      0
#> 196   0.8317734136      0
#> 197  -0.4496361190      0
#> 198   0.1334145338      0
#> 199   2.3071177921      0
#> 200  -0.2294765794      0
#> 201  -0.4651060981      0
#> 202   0.0376596361      0
#> 203   0.1913380254      0
#> 204  -0.5560634732      0
#> 205   1.3994340655      0
#> 206   0.0898730023      0
#> 207   1.2973306856      0
#> 208   1.3480098154      0
#> 209   0.5341298892      0
#> 210  -0.5005423652      0
#> 211  -0.1348763464      0
#> 212  -2.2751023310      0
#> 213  -1.2624788403      0
#> 214  -0.8521772250      0
#> 215   0.9333626881      0
#> 216   0.6534083169      0
#> 217  -0.3664736652      0
#> 218  -0.9796711027      0
#> 219  -0.0875532443      0
#> 220  -0.2464698558      0
#> 221   2.1140914901      0
#> 222  -0.6187784977      0
#> 223   0.2983732678      0
#> 224   0.5032632936      0
#> 225   0.6781473451      0
#> 226  -0.5335590279      0
#> 227  -0.1890926017      0
#> 228  -0.1082040869      0
#> 229   0.1965054027      0
#> 230   0.2924433511      0
#> 231  -1.6780642708      0
#> 232   3.0411050543      0
#> 233  -0.7233213394      0
#> 234  -0.2867963861      0
#> 235   2.6279465901      0
#> 236  -0.1630340311      0
#> 237   0.5955342454      0
#> 238   0.8651201705      0
#> 239   0.5496674604      0
#> 240  -0.2648584234      0
#> 241  -0.1834723403      0
#> 242  -0.3105345200      0
#> 243  -0.2374171915      0
#> 244   0.2397795342      0
#> 245  -0.0437451288      0
#> 246  -0.1242633787      0
#> 247   1.1153746635      0
#> 248   0.5520122580      0
#> 249  -0.8803768241      0
#> 250  -1.9961586243      0
#> 251  -1.0317717618      0
#> 252  -0.7435526497      0
#> 253   1.3080753380      0
#> 254  -0.6765688464      0
#> 255   0.1577107469      0
#> 256  -0.7185460016      0
#> 257  -0.8954995576      0
#> 258  -0.6788381305      0
#> 259  -0.1860912679      0
#> 260  -2.1667484527      0
#> 261   0.1922518946      0
#> 262   0.2195294948      0
#> 263  -0.5720582402      0
#> 264  -1.6236183853      0
#> 265  -1.3080015740      0
#> 266  -0.8409407464      0
#> 267   0.1414441474      0
#> 268   0.5253293753      0
#> 269   0.1582584548      0
#> 270   0.3095142901      0
#> 271   0.6740156803      0
#> 272  -1.2067066005      0
#> 273   0.5159914035      0
#> 274   0.1798777421      0
#> 275  -0.3639208844      0
#> 276  -0.0373077298      0
#> 277  -0.1828207972      0
#> 278  -0.5309081782      0
#> 279   1.6020045559      0
#> 280   0.9123773240      0
#> 281   0.4240622043      0
#> 282  -0.9277381867      0
#> 283   3.0572009746      0
#> 284   0.7162503936      0
#> 285   1.7498592298      0
#> 286  -0.9482685726      0
#> 287   0.3357805865      0
#> 288  -1.0398153331      0
#> 289  -1.8033145832      0
#> 290   1.8687621572      0
#> 291  -1.3539151061      0
#> 292   0.0387382978      0
#> 293  -0.2008515486      0
#> 294   0.0777805089      0
#> 295   0.0663115278      0
#> 296   0.6345120490      0
#> 297   0.6178481550      0
#> 298  -0.3094658498      0
#> 299  -1.1102112619      0
#> 300   0.6109200985      0
#> 301  -1.4807009634      0
#> 302   1.6810202629      0
#> 303  -2.3423283180      0
#> 304  -1.2396531064      0
#> 305  -0.0001976808      0
#> 306  -0.3279903408      0
#> 307  -1.4233128953      0
#> 308   1.4538387761      0
#> 309   1.3687754646      0
#> 310   1.2674098543      0
#> 311   0.0599829753      0
#> 312   0.7335057500      0
#> 313   0.6404748888      0
#> 314  -0.8510995389      0
#> 315  -1.0607168222      0
#> 316  -0.2991621550      0
#> 317   0.7556584525      0
#> 318  -1.8666616150      0
#> 319   1.5575738372      0
#> 320   0.5433260813      0
#> 321   1.0202495452      0
#> 322  -1.3780348114      0
#> 323   0.8150518802      0
#> 324  -0.0290763366      0
#> 325   0.5568937985      0
#> 326  -0.6800638775      0
#> 327   0.8200733303      0
#> 328   0.5214764296      0
#> 329   0.0369148745      0
#> 330   0.5871662802      0
#> 331   0.2429769784      0
#> 332  -1.2109394073      0
#> 333  -1.7004104896      0
#> 334   1.7854454883      0
#> 335  -0.8777673194      0
#> 336   0.8781120565      0
#> 337  -0.5622758550      0
#> 338  -1.3046636267      0
#> 339  -1.6609612493      0
#> 340   1.9622747801      0
#> 341  -1.1084911819      0
#> 342   1.3213416127      0
#> 343   0.0703705753      0
#> 344   0.9461330668      0
#> 345  -0.3226524839      0
#> 346  -0.0782679337      0
#> 347   0.5994108414      0
#> 348   0.0460913835      0
#> 349  -0.9390711794      0
#> 350  -1.7721622486      0
#> 351   0.0383463538      0
#> 352   0.6224863383      0
#> 353   0.4618068082      0
#> 354   1.1387097660      0
#> 355   1.4774782547      0
#> 356  -0.5657613000      0
#> 357  -0.2792995644      0
#> 358   0.1201213489      0
#> 359   0.1750270700      0
#> 360   1.0860350161      0
#> 361  -1.6261657173      0
#> 362   0.6440353590      0
#> 363  -2.7774768258      0
#> 364  -0.8065750686      0
#> 365   0.0180860710      0
#> 366   0.6859398764      0
#> 367  -0.1741255629      0
#> 368   0.2846523976      0
#> 369   1.7029113537      0
#> 370  -0.2967735860      0
#> 371   0.8978487145      0
#> 372   0.7218026270      0
#> 373   0.8452824684      0
#> 374   2.1941742011      0
#> 375  -1.1764603954      0
#> 376   0.5329196183      0
#> 377  -0.2328116883      0
#> 378  -0.3076415749      0
#> 379   0.5804514041      0
#> 380   0.1880793618      0
#> 381   2.0030012367      0
#> 382   0.3242862146      0
#> 383   1.2283873004      0
#> 384  -2.0214834916      0
#> 385   0.1031866026      0
#> 386  -0.5868282099      0
#> 387   0.4149452209      0
#> 388  -0.8572205849      0
#> 389  -1.0104128473      0
#> 390  -0.8934327000      0
#> 391  -0.3126720881      0
#> 392  -1.1430128786      0
#> 393  -1.5609554095      0
#> 394  -1.0224392783      0
#> 395   0.7953192705      0
#> 396   0.4931651777      0
#> 397  -0.3962375981      0
#> 398   1.0009788166      0
#> 399  -1.9825366021      0
#> 400  -0.3142673965      0
#> 401  -1.3033420607      0
#> 402  -0.3949408749      0
#> 403  -0.4947557893      0
#> 404  -0.7264445918      0
#> 405  -0.8805556913      0
#> 406  -2.0955150352      0
#> 407  -0.0014156639      0
#> 408  -0.0638972017      0
#> 409   0.9063865771      0
#> 410   0.5137412159      0
#> 411   1.8911388766      0
#> 412  -0.7072574392      0
#> 413   0.4597654903      0
#> 414   0.1577576565      0
#> 415   0.4456968421      0
#> 416   0.4676828184      0
#> 417  -0.0140931033      0
#> 418   0.7829108170      0
#> 419   0.9519200781      0
#> 420  -0.3487384559      0
#> 421   0.6019179240      0
#> 422   1.1815945644      0
#> 423  -0.3111899544      0
#> 424  -0.3511779445      0
#> 425   0.2545052143      0
#> 426  -0.0179329860      0
#> 427   0.0023578113      0
#> 428  -0.1351595815      0
#> 429   1.5034316468      0
#> 430  -0.7529118698      0
#> 431   0.5629499191      0
#> 432   0.1944878911      0
#> 433   0.5088510752      0
#> 434   0.3958261359      0
#> 435  -1.6570115262      0
#> 436  -1.1417107910      0
#> 437   0.9554204180      0
#> 438  -0.7135820400      0
#> 439  -1.3741742714      0
#> 440  -1.3488007884      0
#> 441   1.2579112330      0
#> 442  -1.2771326835      0
#> 443  -1.0998778235      0
#> 444  -1.0753530506      0
#> 445  -0.6939544027      0
#> 446  -0.0669407835      0
#> 447  -1.4931524661      0
#> 448  -0.0270353947      0
#> 449   0.0005891998      0
#> 450   1.6655296313      0
#> 451  -0.9271455985      0
#> 452  -0.0698301637      0
#> 453   0.3344804475      0
#> 454   0.1318445535      0
#> 455   0.5057897309      0
#> 456   1.6458589464      0
#> 457  -0.3300423568      0
#> 458  -0.2926120458      0
#> 459   0.4203037859      0
#> 460   1.2825686465      0
#> 461  -0.5986524127      0
#> 462   0.7763014348      0
#> 463  -2.9228875706      0
#> 464  -1.2474071507      0
#> 465  -1.6715867192      0
#> 466  -1.2022647222      0
#> 467   0.7489562379      0
#> 468  -0.9463768947      0
#> 469   0.5745621402      0
#> 470   1.2050890411      0
#> 471   0.9361701871      0
#> 472   0.7924229654      0
#> 473  -0.2504795109      0
#> 474   0.6471677681      0
#> 475  -0.9243762968      0
#> 476  -1.1261127920      0
#> 477   1.1279714264      0
#> 478  -2.1169096210      0
#> 479   0.5987043843      0
#> 480   1.2330177157      0
#> 481   0.3954251204      0
#> 482  -1.3090045791      0
#> 483  -0.5169497406      0
#> 484   0.3234953443      0
#> 485   0.4078487051      0
#> 486   0.1088391137      0
#> 487  -1.2638997531      0
#> 488  -1.7843645553      0
#> 489   0.3171006720      0
#> 490   0.4593725140      0
#> 491   1.2203120163      0
#> 492   1.1704372940      0
#> 493  -0.2847380646      0
#> 494   1.1271481601      0
#> 495   0.1685158292      0
#> 496   0.5664023376      0
#> 497   0.0142664174      0
#> 498   0.8686083589      0
#> 499   0.3064980047      0
#> 500  -0.5053200115      0
#> 501   1.0737455883      0
#> 502  -0.2904777898      0
#> 503  -0.4364786858      0
#> 504   0.6319818735      0
#> 505   0.1455548915      0
#> 506   0.2411527694      0
#> 507  -2.2735829793      0
#> 508  -2.2870270235      0
#> 509  -1.1921086907      0
#> 510   0.3735032757      0
#> 511   0.1752999794      0
#> 512  -1.4637245546      0
#> 513   0.6789638116      0
#> 514   1.0789394875      0
#> 515  -0.5615444324      0
#> 516  -0.6135631258      0
#> 517  -0.0437333072      0
#> 518  -0.8980729896      0
#> 519   1.5572076172      0
#> 520  -0.0017608275      0
#> 521   0.6016372051      0
#> 522  -1.5578337240      0
#> 523  -0.7016834517      0
#> 524  -1.1750199518      0
#> 525  -0.6799735669      0
#> 526  -1.3768272668      0
#> 527   1.2449718812      0
#> 528  -0.7771064484      0
#> 529  -0.4912294425      0
#> 530  -0.4973614701      0
#> 531   1.1882887341      0
#> 532  -0.8805552419      0
#> 533  -1.5427181338      0
#> 534   1.7431917263      0
#> 535   1.7988984349      0
#> 536  -0.9754796936      0
#> 537   1.3081704906      0
#> 538  -0.8393263467      0
#> 539  -0.8069086762      0
#> 540  -1.1119450225      0
#> 541  -0.8264795452      0
#> 542   0.4978003581      0
#> 543  -0.7617589333      0
#> 544   1.3571442050      0
#> 545  -1.5951244199      0
#> 546   0.5041294411      0
#> 547  -1.0905450141      0
#> 548  -0.7699174867      0
#> 549  -0.4256412119      0
#> 550   0.4779677227      0
#> 551   0.3817402360      0
#> 552  -0.4364599260      0
#> 553  -0.8270103571      0
#> 554  -0.9482946328      0
#> 555  -0.2651489879      0
#> 556   0.6262018088      0
#> 557  -1.8791722717      0
#> 558   1.5797350688      0
#> 559  -1.1038457264      0
#> 560   0.4548438080      0
#> 561   0.0112445114      0
#> 562  -0.2734455161      0
#> 563   0.1905850326      0
#> 564  -0.1823094638      0
#> 565  -0.6614342797      0
#> 566  -0.7101773429      0
#> 567   0.0605877957      0
#> 568  -0.6975025435      0
#> 569   0.1816193946      0
#> 570   0.7005149182      0
#> 571  -1.2026292109      0
#> 572   1.1896620776      0
#> 573   0.7601190954      0
#> 574  -1.9422433848      0
#> 575   1.4566006726      0
#> 576  -0.0925634056      0
#> 577   0.3324169895      0
#> 578  -0.9702623256      0
#> 579   0.6886958352      0
#> 580   1.5672615215      0
#> 581  -1.3261018375      0
#> 582   0.1620453584      0
#> 583   0.1893087873      0
#> 584   0.5298438505      0
#> 585  -1.2070496963      0
#> 586  -0.8026434807      0
#> 587   1.1249600401      0
#> 588  -1.4126118455      0
#> 589  -0.0266712079      0
#> 590  -0.8424308036      0
#> 591   1.8371144809      0
#> 592  -1.6317577366      0
#> 593  -1.3503068079      0
#> 594  -0.5983596318      0
#> 595  -0.9786751549      0
#> 596  -0.2006944782      0
#> 597   1.0130720450      0
#> 598  -0.9423935877      0
#> 599   0.2546653407      0
#> 600  -0.1057735458      0
#> 601   0.6652828019      0
#> 602  -1.8776505366      0
#> 603   1.0981386909      0
#> 604   1.7619463943      0
#> 605   0.7189216104      0
#> 606  -0.0173030028      0
#> 607   2.1891055634      0
#> 608  -0.4530359359      0
#> 609  -0.6896477202      0
#> 610   0.3081606822      0
#> 611  -0.4908665687      0
#> 612  -0.4375131974      0
#> 613  -1.0757657655      0
#> 614   0.0967754474      0
#> 615  -0.1128446870      0
#> 616   0.3472027434      0
#> 617  -0.1603861831      0
#> 618   1.0921974013      0
#> 619  -0.0154217202      0
#> 620  -0.5481639111      0
#> 621   0.0589746039      0
#> 622   1.5497158747      0
#> 623  -1.4518035672      0
#> 624  -0.5192770351      0
#> 625   0.3505224773      0
#> 626  -0.1728278925      0
#> 627   0.5094170445      0
#> 628  -0.2312087974      0
#> 629   1.7320617282      0
#> 630   0.2875673033      0
#> 631  -0.4645938538      0
#> 632   0.1758294850      0
#> 633  -0.8321770774      0
#> 634  -0.2503320839      0
#> 635   0.4942369782      0
#> 636  -0.4461579417      0
#> 637   0.1291736106      0
#> 638   0.6852922432      0
#> 639   2.6103356541      0
#> 640  -0.6117284476      0
#> 641   0.0427642322      0
#> 642   0.5020400474      0
#> 643  -1.4651300346      0
#> 644  -0.3601803487      0
#> 645   1.8661697201      0
#> 646  -2.3291822875      0
#> 647  -1.6924859984      0
#> 648   1.3138949017      0
#> 649  -0.2792457951      0
#> 650  -0.4499292753      0
#> 651  -0.6467362217      0
#> 652   0.6441402489      0
#> 653   0.4284634600      0
#> 654  -2.6018667253      0
#> 655   1.3852939561      0
#> 656  -0.5001203952      0
#> 657   0.8297760687      0
#> 658   0.7823933639      0
#> 659   1.4874934636      0
#> 660  -1.9276846349      0
#> 661   0.9324027620      0
#> 662  -0.2401672894      0
#> 663   0.5222961224      0
#> 664   0.9639855057      0
#> 665   2.8130697789      0
#> 666  -0.6767460868      0
#> 667  -0.6236535612      0
#> 668  -0.4373905510      0
#> 669  -1.2193965029      0
#> 670   0.9659066793      0
#> 671  -0.0358995951      0
#> 672   0.1552389485      0
#> 673  -1.5068581731      0
#> 674  -0.1769239518      0
#> 675  -0.3425207496      0
#> 676   0.2028652497      0
#> 677   0.3019001324      0
#> 678  -0.9554091892      0
#> 679  -1.2020301243      0
#> 680  -0.4511490611      0
#> 681   1.6396806068      0
#> 682   2.0674311716      0
#> 683   0.2881023839      0
#> 684  -1.1658250553      0
#> 685   0.4635644919      0
#> 686   1.3661758145      0
#> 687  -0.4385676925      0
#> 688  -1.9873330939      0
#> 689   1.7716134688      0
#> 690   0.1724210803      0
#> 691   1.3187988243      0
#> 692  -0.0736626879      0
#> 693  -1.9356814504      0
#> 694  -1.2085632244      0
#> 695   0.4781074463      0
#> 696   0.4762522637      0
#> 697  -1.4354081876      0
#> 698   1.3415222885      0
#> 699  -0.6878214145      0
#> 700   1.1410588843      0
#> 701  -0.5443186363      0
#> 702   0.2576003883      0
#> 703   2.3958915151      0
#> 704  -0.8589790927      0
#> 705  -0.5476116753      0
#> 706   1.7407450628      0
#> 707   0.7899912841      0
#> 708   0.6013830615      0
#> 709  -0.5403728822      0
#> 710  -0.1881628484      0
#> 711  -1.0097961153      0
#> 712  -0.3654809436      0
#> 713  -0.0886850726      0
#> 714  -0.4333727558      0
#> 715  -1.0028013429      0
#> 716  -0.4781739742      0
#> 717   0.1825992899      0
#> 718  -1.8131059253      0
#> 719   0.1987050365      0
#> 720  -0.8608749487      0
#> 721  -1.1246981562      0
#> 722   0.2113823247      0
#> 723   0.9765630383      0
#> 724   0.5986744865      0
#> 725  -0.9305339000      0
#> 726  -0.9296887878      0
#> 727   0.0092541396      0
#> 728   0.2768778025      0
#> 729   0.7104454351      0
#> 730  -0.0777913351      0
#> 731   1.3874245348      0
#> 732  -1.0591739393      0
#> 733   0.4406811503      0
#> 734  -0.7798252857      0
#> 735  -0.3473723994      0
#> 736   0.6735440278      0
#> 737   1.5859853170      0
#> 738  -0.4217916148      0
#> 739   0.9194867788      0
#> 740  -1.5509458710      0
#> 741   0.9552484558      0
#> 742  -0.3865176066      0
#> 743   1.2329706431      0
#> 744   1.2807122531      0
#> 745  -0.9942001591      0
#> 746   1.0753874843      0
#> 747   0.6178662414      0
#> 748  -0.5345064635      0
#> 749   0.4260288440      0
#> 750  -0.7152158100      0
#> 751  -1.6995731039      0
#> 752  -1.9726900457      0
#> 753  -1.0401642334      0
#> 754   0.0467297732      0
#> 755   1.7507834600      0
#> 756   0.3218667587      0
#> 757  -0.8937065955      0
#> 758  -0.9939159651      0
#> 759  -0.4317736468      0
#> 760   0.3720963231      0
#> 761  -0.2840231978      0
#> 762   0.0208069888      0
#> 763   0.1998402325      0
#> 764  -1.0210263256      0
#> 765  -0.2667313836      0
#> 766   0.6089838685      0
#> 767   0.0991546507      0
#> 768   0.9058050572      0
#> 769   0.5318384536      0
#> 770  -1.2556113223      0
#> 771   0.4330639693      0
#> 772  -0.7388577020      0
#> 773   0.0931431645      0
#> 774   0.8583040584      0
#> 775   1.0136176496      0
#> 776  -1.0690658823      0
#> 777  -0.8094113889      0
#> 778   1.5061425179      0
#> 779  -0.3784578384      0
#> 780  -0.4873789383      0
#> 781  -1.0600947513      0
#> 782   0.9245566243      0
#> 783  -2.2468734933      0
#> 784  -0.2328601189      0
#> 785  -0.1684436591      0
#> 786  -0.4013049036      0
#> 787   2.0500377538      0
#> 788   1.1517763759      0
#> 789  -1.9785752304      0
#> 790  -0.8791815757      0
#> 791   0.0697266467      0
#> 792   1.7738630401      0
#> 793   1.3865866838      0
#> 794   0.7054353267      0
#> 795   0.1619938875      0
#> 796   1.4687659700      0
#> 797  -0.6320329099      0
#> 798  -0.2598737944      0
#> 799  -0.1043062790      0
#> 800  -0.0957066266      0
#> 801  -0.6939818912      0
#> 802  -1.2300366330      0
#> 803   1.1729802286      0
#> 804  -1.6968315951      0
#> 805   1.0658724349      0
#> 806   0.0193836055      0
#> 807   0.0598763189      0
#> 808  -1.0197934230      0
#> 809  -0.5663426357      0
#> 810   0.8235436293      0
#> 811  -1.1106717013      0
#> 812   1.1274895640      0
#> 813  -0.0816751745      0
#> 814   0.6089298553      0
#> 815   0.6414187946      0
#> 816  -0.6378329586      0
#> 817  -1.3317913292      0
#> 818  -0.9365138248      0
#> 819   0.9891728967      0
#> 820  -0.3841174868      0
#> 821   1.8596203673      0
#> 822  -1.5527913465      0
#> 823   1.8233546713      0
#> 824  -0.3616632514      0
#> 825   0.3583962686      0
#> 826   0.9472283982      0
#> 827  -0.1413188043      0
#> 828  -1.4522111075      0
#> 829  -0.2220994113      0
#> 830  -2.1874087016      0
#> 831   0.9292875432      0
#> 832  -0.3898616132      0
#> 833  -0.6242925770      0
#> 834  -1.3917017775      0
#> 835   0.5565259381      0
#> 836  -1.9994211321      0
#> 837  -0.8855311625      0
#> 838   0.7285073278      0
#> 839   0.0309404658      0
#> 840   0.3020491128      0
#> 841  -0.4683966577      0
#> 842   0.6066465752      0
#> 843  -0.8519912787      0
#> 844  -0.9146908548      0
#> 845  -0.2990397430      0
#> 846  -1.5492298598      0
#> 847  -0.1204170178      0
#> 848   1.0658914897      0
#> 849  -1.4116160741      0
#> 850  -0.1194423592      0
#> 851   1.3622464243      0
#> 852   0.1917030609      0
#> 853  -0.6960506269      0
#> 854  -1.4466953152      0
#> 855   1.9374549236      0
#> 856   0.6540059313      0
#> 857  -2.1931607159      0
#> 858   1.2906209250      0
#> 859  -1.5505069041      0
#> 860   0.4578654231      0
#> 861  -0.2993848675      0
#> 862  -0.7538721521      0
#> 863   0.2879378103      0
#> 864   0.4340591033      0
#> 865  -0.5234511307      0
#> 866  -0.9772120700      0
#> 867   0.2998360092      0
#> 868  -0.2392409077      0
#> 869   1.0720793132      0
#> 870   0.8671196863      0
#> 871  -0.4004174329      0
#> 872   1.4907753657      0
#> 873  -1.8434346722      0
#> 874  -0.0745618691      0
#> 875  -0.2672044677      0
#> 876  -0.2470320368      0
#> 877  -0.6703638134      0
#> 878   1.7174345327      0
#> 879  -0.4218942773      0
#> 880  -1.8153673040      0
#> 881   0.3858238955      0
#> 882  -0.7459049923      0
#> 883  -1.8585080429      0
#> 884  -1.3135676998      0
#> 885   0.5068432065      0
#> 886   1.3494195389      0
#> 887   0.6139819083      0
#> 888   1.1245525412      0
#> 889   0.1126200514      0
#> 890   0.1544749176      0
#> 891  -0.3841358553      0
#> 892  -1.4311600315      0
#> 893  -1.6028868832      0
#> 894  -0.9823338267      0
#> 895   0.4574593709      0
#> 896   0.8319026788      0
#> 897  -0.2859247917      0
#> 898  -1.3977319598      0
#> 899   0.2276365636      0
#> 900   1.2667882099      1
#> 901   1.3011292225      1
#> 902   0.6975073133      1
#> 903   1.7062249618      1
#> 904   0.5193317002      1
#> 905   1.9913126255      1
#> 906   0.6375768448      1
#> 907  -0.6294436235      1
#> 908   0.6466840399      1
#> 909   1.1344416011      1
#> 910   1.1065783968      1
#> 911   1.9651267954      1
#> 912  -0.2078224454      1
#> 913   0.2277245960      1
#> 914   2.4627120908      1
#> 915   1.3984955603      1
#> 916   1.1699831041      1
#> 917  -0.6396294362      1
#> 918  -0.4678519530      1
#> 919   2.9137525301      1
#> 920   0.6082009327      1
#> 921   0.3095636116      1
#> 922   1.8422985576      1
#> 923  -0.7664264822      1
#> 924  -0.3702423028      1
#> 925  -0.1319144521      1
#> 926  -0.0525165497      1
#> 927   0.1584130098      1
#> 928  -0.2234585900      1
#> 929  -0.4375789283      1
#> 930  -0.9248467566      1
#> 931   2.0481873625      1
#> 932   0.6280678839      1
#> 933   1.3264414979      1
#> 934  -0.1232729828      1
#> 935   1.1189030363      1
#> 936   0.3910236663      1
#> 937   1.2331056611      1
#> 938   1.5508326600      1
#> 939   2.2567604811      1
#> 940   2.0556210681      1
#> 941   2.1861851088      1
#> 942   0.3861929341      1
#> 943   0.7237481402      1
#> 944   1.8724008063      1
#> 945   1.3350371369      1
#> 946   0.7307271168      1
#> 947  -0.4290796165      1
#> 948   0.6483819149      1
#> 949   2.3676045795      1
#> 950  -0.1943536164      1
#> 951  -0.4618159976      1
#> 952   1.6783298527      1
#> 953   1.9835228608      1
#> 954  -1.2220397810      1
#> 955  -0.0409938240      1
#> 956   1.3047066209      1
#> 957  -0.1492794569      1
#> 958  -0.1160744426      1
#> 959  -0.3788576634      1
#> 960  -0.2201459611      1
#> 961   0.3706608437      1
#> 962   1.7750506694      1
#> 963   1.6220196806      1
#> 964   2.2959245242      1
#> 965   0.3291753555      1
#> 966   0.8602170119      1
#> 967   1.6322574000      1
#> 968   0.1956240660      1
#> 969   0.6625830830      1
#> 970  -0.2268891893      1
#> 971   0.1602000049      1
#> 972   1.0265925214      1
#> 973   0.4274418673      1
#> 974   1.9630239790      1
#> 975   2.4502011550      1
#> 976  -0.4153540202      1
#> 977   1.0028943324      1
#> 978   0.0126103329      1
#> 979   0.3511199322      1
#> 980   1.9295555125      1
#> 981   0.6521912848      1
#> 982   0.1900181726      1
#> 983   0.1618499520      1
#> 984   0.7479095658      1
#> 985   1.9196181741      1
#> 986  -1.1243928743      1
#> 987   0.6754524386      1
#> 988   2.5053142843      1
#> 989   1.1867880426      1
#> 990  -0.4621280722      1
#> 991   1.9128910714      1
#> 992   1.4519140902      1
#> 993   0.7855683201      1
#> 994  -0.1591796639      1
#> 995  -0.8872816379      1
#> 996  -0.8047678251      1
#> 997   0.1674562411      1
#> 998  -0.9184315814      1
#> 999  -0.1188223097      1
#> 1000  0.7480745352      1