Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.208032230      0
#> 2    -0.358536285      0
#> 3     0.240955448      0
#> 4     0.782106845      0
#> 5     0.138633073      0
#> 6    -1.695126211      0
#> 7    -0.586573930      0
#> 8    -0.918820902      0
#> 9    -0.136875008      0
#> 10   -0.025639543      0
#> 11   -2.038870640      0
#> 12   -0.464129768      0
#> 13    0.890299754      0
#> 14    1.148772568      0
#> 15    0.754331869      0
#> 16   -0.528540927      0
#> 17    0.151363218      0
#> 18    0.074527680      0
#> 19    0.517208129      0
#> 20   -0.223777543      0
#> 21    0.113576904      0
#> 22    0.205078542      0
#> 23   -0.100577459      0
#> 24    1.979594360      0
#> 25    1.311293761      0
#> 26   -1.058922541      0
#> 27   -1.267302851      0
#> 28    0.333005293      0
#> 29   -0.107141162      0
#> 30    0.615966482      0
#> 31   -1.736238392      0
#> 32    0.157403244      0
#> 33    0.685992095      0
#> 34   -0.404443895      0
#> 35    0.662985919      0
#> 36   -1.126791800      0
#> 37    1.019030844      0
#> 38   -1.251774870      0
#> 39    0.531873688      0
#> 40    0.279951485      0
#> 41    0.059033967      0
#> 42    1.567661489      0
#> 43   -0.958980353      0
#> 44   -0.678705267      0
#> 45    1.040646475      0
#> 46    0.304094031      0
#> 47    0.564896880      0
#> 48   -0.689998406      0
#> 49   -0.082677312      0
#> 50   -0.961386755      0
#> 51    0.148388302      0
#> 52   -0.463162849      0
#> 53    2.110818734      0
#> 54   -0.796666959      0
#> 55   -0.342813658      0
#> 56   -2.371447944      0
#> 57   -0.863873442      0
#> 58    1.556456251      0
#> 59    0.400129573      0
#> 60    0.150118236      0
#> 61   -0.103863660      0
#> 62   -0.341582993      0
#> 63    0.059581437      0
#> 64   -1.467154846      0
#> 65    1.073410228      0
#> 66    0.781001828      0
#> 67    0.256992213      0
#> 68    0.891439769      0
#> 69   -1.145234583      0
#> 70    0.657236893      0
#> 71    0.905818463      0
#> 72   -0.018273896      0
#> 73    0.213492160      0
#> 74    0.210248387      0
#> 75   -0.980387876      0
#> 76   -1.438325024      0
#> 77    0.100273096      0
#> 78   -0.636830357      0
#> 79    0.355451715      0
#> 80    0.015997289      0
#> 81    0.216667103      0
#> 82   -0.977445105      0
#> 83   -0.788637419      0
#> 84   -0.130607229      0
#> 85    0.118677856      0
#> 86   -1.369744807      0
#> 87   -0.699801182      0
#> 88   -1.131419809      0
#> 89    0.083576897      0
#> 90    1.355474668      0
#> 91   -0.729084151      0
#> 92    1.069306447      0
#> 93    1.655124734      0
#> 94    1.027393937      0
#> 95    0.689590704      0
#> 96   -1.793587478      0
#> 97    2.639894205      0
#> 98   -0.472849305      0
#> 99   -0.402197223      0
#> 100  -0.557357606      0
#> 101   0.369020290      0
#> 102   0.331655443      0
#> 103   0.294263042      0
#> 104  -2.172411057      0
#> 105   0.514771166      0
#> 106  -0.494491231      0
#> 107  -0.953398914      0
#> 108  -0.225745446      0
#> 109   0.284502030      0
#> 110  -0.623705403      0
#> 111   0.143276576      0
#> 112   1.334861543      0
#> 113  -1.500943547      0
#> 114   0.013651167      0
#> 115   1.945307981      0
#> 116  -0.433323661      0
#> 117  -1.729039197      0
#> 118   0.298987297      0
#> 119  -0.364178222      0
#> 120   0.041641797      0
#> 121  -0.463957245      0
#> 122   0.318336308      0
#> 123   0.125558487      0
#> 124  -0.685791677      0
#> 125  -1.836715498      0
#> 126   1.523356906      0
#> 127   0.071152932      0
#> 128  -0.264184731      0
#> 129   0.925119184      0
#> 130  -0.566707292      0
#> 131   0.176513962      0
#> 132  -0.338262351      0
#> 133  -1.057439602      0
#> 134   0.490591076      0
#> 135   1.773659354      0
#> 136   1.658950814      0
#> 137  -0.403930759      0
#> 138   0.255806484      0
#> 139   2.015350256      0
#> 140   1.916007166      0
#> 141  -0.602206846      0
#> 142   0.057947259      0
#> 143   1.384892951      0
#> 144   0.924330733      0
#> 145  -1.172730005      0
#> 146  -0.046260273      0
#> 147   0.650094499      0
#> 148   0.188425266      0
#> 149  -0.013618846      0
#> 150   0.331614539      0
#> 151   0.097493977      0
#> 152  -0.847211171      0
#> 153   1.639981292      0
#> 154   0.017673246      0
#> 155   0.246600616      0
#> 156  -0.917643821      0
#> 157   1.848968843      0
#> 158   0.275504548      0
#> 159   1.084013996      0
#> 160   0.664910003      0
#> 161   0.073468784      0
#> 162   2.279743446      0
#> 163   0.424927005      0
#> 164  -1.658277912      0
#> 165   1.292286736      0
#> 166  -1.843053714      0
#> 167  -1.723610528      0
#> 168   0.806846336      0
#> 169   0.482239351      0
#> 170   0.208859742      0
#> 171   0.942929764      0
#> 172   0.413459158      0
#> 173   0.071260388      0
#> 174  -0.666042430      0
#> 175   0.200966547      0
#> 176   1.165681995      0
#> 177  -1.543954863      0
#> 178  -0.960520081      0
#> 179  -0.778495926      0
#> 180   0.455164574      0
#> 181  -0.554336684      0
#> 182  -1.275089334      0
#> 183  -0.924845111      0
#> 184  -1.441318455      0
#> 185  -0.001672856      0
#> 186  -0.254618894      0
#> 187  -0.654426062      0
#> 188  -1.432130414      0
#> 189   0.917868732      0
#> 190   0.603885985      0
#> 191   0.641814450      0
#> 192   0.635851990      0
#> 193   0.043197385      0
#> 194   0.872040032      0
#> 195   0.949198067      0
#> 196  -0.075534786      0
#> 197  -1.504466619      0
#> 198  -1.274202382      0
#> 199   0.939384401      0
#> 200   0.412972057      0
#> 201  -1.459411768      0
#> 202  -1.157023746      0
#> 203  -1.876607638      0
#> 204  -1.816257783      0
#> 205   0.711701833      0
#> 206  -1.402773534      0
#> 207   0.712349655      0
#> 208  -3.284727263      0
#> 209   0.127501960      0
#> 210  -1.059893292      0
#> 211  -1.513485450      0
#> 212  -0.585480464      0
#> 213  -1.266807935      0
#> 214   0.054467859      0
#> 215  -1.024756772      0
#> 216  -0.800593193      0
#> 217   0.392463077      0
#> 218  -1.409037461      0
#> 219   1.141803690      0
#> 220  -0.876240322      0
#> 221  -0.039284566      0
#> 222  -0.240759535      0
#> 223   0.565922506      0
#> 224  -0.425193207      0
#> 225   0.170993496      0
#> 226  -0.468855409      0
#> 227   0.025858738      0
#> 228   0.015469571      0
#> 229  -0.636633169      0
#> 230   0.835701841      0
#> 231   0.206712915      0
#> 232   0.047981502      0
#> 233  -0.680429237      0
#> 234  -0.706961630      0
#> 235   0.928208799      0
#> 236  -0.122904664      0
#> 237   1.352675755      0
#> 238   1.453665489      0
#> 239  -0.507149954      0
#> 240  -0.690921509      0
#> 241   0.713659272      0
#> 242  -1.435494778      0
#> 243  -0.068267863      0
#> 244  -0.735244404      0
#> 245  -0.178827500      0
#> 246  -1.853916003      0
#> 247  -1.915930262      0
#> 248  -1.345165646      0
#> 249   0.142963789      0
#> 250   0.227382887      0
#> 251   0.916599371      0
#> 252  -1.606153932      0
#> 253  -0.584788251      0
#> 254   0.695282452      0
#> 255  -1.978972532      0
#> 256  -1.514426127      0
#> 257  -0.691496694      0
#> 258   1.354619638      0
#> 259  -1.115016226      0
#> 260   0.294929824      0
#> 261  -1.585957933      0
#> 262   0.099775933      0
#> 263   1.190281015      0
#> 264  -0.282111324      0
#> 265  -1.024997952      0
#> 266  -0.100055643      0
#> 267   0.324364011      0
#> 268   0.551632676      0
#> 269  -0.927172004      0
#> 270   0.573503922      0
#> 271   0.508515174      0
#> 272  -0.038067391      0
#> 273   0.785297458      0
#> 274  -2.611881584      0
#> 275  -0.033250026      0
#> 276   1.319141442      0
#> 277  -0.752639036      0
#> 278  -1.148395386      0
#> 279   2.020818308      0
#> 280   0.279800171      0
#> 281   0.647874269      0
#> 282   0.922672775      0
#> 283   0.508513106      0
#> 284   1.009855499      0
#> 285  -0.218606739      0
#> 286   0.128124881      0
#> 287  -0.115692434      0
#> 288   1.626193710      0
#> 289   2.601085986      0
#> 290  -0.664745831      0
#> 291   0.138236302      0
#> 292  -0.507197459      0
#> 293  -0.245353373      0
#> 294   0.341580292      0
#> 295   1.087156646      0
#> 296  -1.327927353      0
#> 297  -1.017396910      0
#> 298  -1.158003335      0
#> 299  -1.035170677      0
#> 300   1.815022397      0
#> 301  -0.264481555      0
#> 302  -0.315036143      0
#> 303  -0.358335577      0
#> 304  -0.489064559      0
#> 305  -0.976803492      0
#> 306  -0.972149128      0
#> 307  -0.583203621      0
#> 308   0.293741781      0
#> 309  -0.148388478      0
#> 310  -0.456800747      0
#> 311  -1.076137209      0
#> 312  -0.201412784      0
#> 313   0.831991163      0
#> 314   0.186478349      0
#> 315   1.095676377      0
#> 316  -0.685392934      0
#> 317   1.265294312      0
#> 318   0.496148030      0
#> 319  -0.905854428      0
#> 320   0.323240282      0
#> 321   0.365035794      0
#> 322  -0.022537006      0
#> 323   1.050626272      0
#> 324   1.250283162      0
#> 325   0.405793712      0
#> 326   0.118082110      0
#> 327   0.038144506      0
#> 328   1.160949459      0
#> 329   1.461100256      0
#> 330  -0.748437776      0
#> 331   0.637206271      0
#> 332   1.316345101      0
#> 333   0.101518548      0
#> 334   0.447460783      0
#> 335   1.061288858      0
#> 336  -2.870376168      0
#> 337  -0.072603228      0
#> 338   1.262889008      0
#> 339   0.045990128      0
#> 340  -0.361207293      0
#> 341   0.570720796      0
#> 342   0.871679188      0
#> 343  -0.440183228      0
#> 344  -0.206705888      0
#> 345   1.114908052      0
#> 346  -1.063123401      0
#> 347   0.193076430      0
#> 348  -0.543865203      0
#> 349   1.065422701      0
#> 350   0.285507940      0
#> 351  -0.691339950      0
#> 352  -0.107755632      0
#> 353  -0.111508133      0
#> 354   0.675989503      0
#> 355   0.493214210      0
#> 356   0.725673342      0
#> 357  -0.330131671      0
#> 358  -1.050391364      0
#> 359  -0.441488014      0
#> 360  -0.984378179      0
#> 361  -0.827431519      0
#> 362  -0.083492246      0
#> 363  -1.787387407      0
#> 364  -0.900766038      0
#> 365   0.157371317      0
#> 366   1.239229754      0
#> 367   0.614666029      0
#> 368   1.205750508      0
#> 369   1.290039497      0
#> 370   0.119849766      0
#> 371   0.077506953      0
#> 372   0.590656925      0
#> 373   1.141975321      0
#> 374   0.421660738      0
#> 375  -2.122369396      0
#> 376  -0.686165136      0
#> 377  -1.351918016      0
#> 378   0.692690209      0
#> 379  -0.189710488      0
#> 380   0.587918343      0
#> 381  -0.828554568      0
#> 382  -1.353397625      0
#> 383  -0.329201970      0
#> 384  -1.640691251      0
#> 385  -1.271675241      0
#> 386  -2.347703781      0
#> 387   0.743452486      0
#> 388  -0.433428785      0
#> 389   2.311695154      0
#> 390   1.409017074      0
#> 391   0.980557603      0
#> 392  -0.761941460      0
#> 393  -0.849499750      0
#> 394   0.906362617      0
#> 395   0.315699000      0
#> 396  -0.063208523      0
#> 397  -1.010825713      0
#> 398   3.040363461      0
#> 399  -0.468055959      0
#> 400  -1.325146216      0
#> 401   1.143169429      0
#> 402   1.518249415      0
#> 403   0.602123064      0
#> 404   0.080661526      0
#> 405   0.045281144      0
#> 406  -0.195965234      0
#> 407  -0.135470451      0
#> 408  -0.565254160      0
#> 409   0.565486871      0
#> 410   0.255416099      0
#> 411   0.481768856      0
#> 412  -0.956686920      0
#> 413  -0.185609717      0
#> 414   0.621671830      0
#> 415  -3.034119471      0
#> 416  -0.287018022      0
#> 417  -1.240990330      0
#> 418   0.490215047      0
#> 419   0.273702205      0
#> 420  -0.226155975      0
#> 421   0.525709605      0
#> 422   1.441664772      0
#> 423  -1.026587346      0
#> 424   1.452685603      0
#> 425  -1.744952654      0
#> 426   1.219724283      0
#> 427   0.021407468      0
#> 428   0.046273630      0
#> 429   1.186930907      0
#> 430   0.815484063      0
#> 431   2.122427765      0
#> 432  -0.175373418      0
#> 433  -1.336954372      0
#> 434  -0.594251526      0
#> 435  -1.263695749      0
#> 436  -0.338262515      0
#> 437   0.242423959      0
#> 438   0.845293393      0
#> 439   1.172776731      0
#> 440   0.470114418      0
#> 441   0.614133315      0
#> 442  -0.151201727      0
#> 443  -1.400249854      0
#> 444  -0.312214889      0
#> 445   0.505607523      0
#> 446   0.511322623      0
#> 447  -1.242804818      0
#> 448   0.294185388      0
#> 449  -0.807511197      0
#> 450  -0.119060715      0
#> 451  -0.734570245      0
#> 452  -0.167715952      0
#> 453  -0.394578298      0
#> 454   0.145696800      0
#> 455  -0.855494911      0
#> 456  -0.177727672      0
#> 457  -0.027506205      0
#> 458  -1.065070788      0
#> 459  -0.547541137      0
#> 460  -1.526807891      0
#> 461  -0.281771597      0
#> 462  -0.076968454      0
#> 463   0.178025949      0
#> 464   0.282012326      0
#> 465  -1.059870594      0
#> 466   0.767178268      0
#> 467  -1.486359668      0
#> 468  -0.540876747      0
#> 469   0.787991279      0
#> 470   0.355552036      0
#> 471   1.822832629      0
#> 472  -0.374338223      0
#> 473   0.522990613      0
#> 474  -0.258562039      0
#> 475  -0.836726473      0
#> 476  -0.318801189      0
#> 477   1.037924363      0
#> 478  -0.289720712      0
#> 479   2.285025395      0
#> 480   0.039802014      0
#> 481   1.021329901      0
#> 482   0.563818444      0
#> 483  -0.321806640      0
#> 484   1.236938149      0
#> 485  -0.340596992      0
#> 486   1.087855581      0
#> 487  -0.747276117      0
#> 488  -1.415401595      0
#> 489  -0.523040643      0
#> 490  -0.260519296      0
#> 491  -0.757046143      0
#> 492  -1.335967745      0
#> 493   1.674976772      0
#> 494   1.350359099      0
#> 495  -0.683178471      0
#> 496   0.229578640      0
#> 497  -2.161619538      0
#> 498  -1.614762122      0
#> 499   0.374808535      0
#> 500   0.929141430      0
#> 501  -1.032676312      0
#> 502   0.808768104      0
#> 503   2.232070597      0
#> 504   1.627433109      0
#> 505  -0.988949694      0
#> 506  -1.432308257      0
#> 507   1.368492713      0
#> 508   0.260240192      0
#> 509  -0.016772712      0
#> 510   0.053972107      0
#> 511  -0.394521778      0
#> 512   1.027699766      0
#> 513  -2.001880606      0
#> 514  -0.252534947      0
#> 515   0.319903818      0
#> 516   2.569140534      0
#> 517  -0.492904456      0
#> 518  -0.171428397      0
#> 519  -0.335988696      0
#> 520   1.812293317      0
#> 521  -0.291207740      0
#> 522   0.784157272      0
#> 523   1.217748492      0
#> 524  -0.146740923      0
#> 525   0.207544064      0
#> 526  -1.649432105      0
#> 527   0.945351287      0
#> 528  -0.297050893      0
#> 529   0.136003454      0
#> 530   0.414120661      0
#> 531  -0.007934384      0
#> 532  -0.862534019      0
#> 533  -1.064864727      0
#> 534   1.536314999      0
#> 535   0.453450176      0
#> 536   0.193593098      0
#> 537   0.806135443      0
#> 538   0.327617122      0
#> 539  -0.696763136      0
#> 540   0.977718264      0
#> 541  -0.231728922      0
#> 542  -0.230718874      0
#> 543   2.283828239      0
#> 544  -0.235791105      0
#> 545   1.385514568      0
#> 546  -0.291579295      0
#> 547  -0.312774137      0
#> 548  -1.589960757      0
#> 549   1.875477890      0
#> 550  -0.095208590      0
#> 551  -0.444766457      0
#> 552   0.516679932      0
#> 553   0.498956161      0
#> 554   0.537446513      0
#> 555  -0.545135724      0
#> 556  -0.295745587      0
#> 557  -0.399519928      0
#> 558   0.210154941      0
#> 559  -0.158808364      0
#> 560  -0.302693726      0
#> 561   1.862129432      0
#> 562   0.426584512      0
#> 563  -0.124056188      0
#> 564  -1.926853637      0
#> 565   1.912001933      0
#> 566   0.699444237      0
#> 567  -0.994982976      0
#> 568   0.873964937      0
#> 569   0.934975890      0
#> 570   0.186210560      0
#> 571   1.177231793      0
#> 572   0.021070104      0
#> 573   0.697281726      0
#> 574   0.062984182      0
#> 575   0.332525903      0
#> 576   0.214463087      0
#> 577  -1.907561426      0
#> 578  -0.809268680      0
#> 579  -0.641332336      0
#> 580   0.600604953      0
#> 581   0.217898361      0
#> 582   0.722310064      0
#> 583   1.011900833      0
#> 584  -0.051022341      0
#> 585   0.631973806      0
#> 586  -0.609338538      0
#> 587   0.821412719      0
#> 588  -2.088830168      0
#> 589  -1.212534474      0
#> 590   0.655780147      0
#> 591   0.519037305      0
#> 592   0.773061588      0
#> 593  -0.873702198      0
#> 594  -0.231081543      0
#> 595   0.916660134      0
#> 596  -2.351365276      0
#> 597  -0.647543451      0
#> 598  -0.513817772      0
#> 599   0.057822298      0
#> 600   2.681653643      0
#> 601   1.354942521      0
#> 602  -0.982841996      0
#> 603  -1.363105745      0
#> 604   0.423318159      0
#> 605   2.041909832      0
#> 606   1.258049960      0
#> 607   0.801121304      0
#> 608   0.952379168      0
#> 609  -1.005684437      0
#> 610   1.361944200      0
#> 611   0.397185640      0
#> 612   1.422131028      0
#> 613   0.396052311      0
#> 614   0.656659210      0
#> 615   0.675034078      0
#> 616  -0.388086381      0
#> 617   0.664093802      0
#> 618  -0.403230175      0
#> 619   0.448945618      0
#> 620  -0.910507616      0
#> 621  -0.630659160      0
#> 622  -1.132354092      0
#> 623  -1.989808819      0
#> 624  -1.086933670      0
#> 625   0.562221984      0
#> 626   1.165080948      0
#> 627   1.367037213      0
#> 628   1.194320282      0
#> 629   1.633937164      0
#> 630  -0.498856221      0
#> 631   0.711562057      0
#> 632  -0.153316642      0
#> 633   1.111828101      0
#> 634   1.046075173      0
#> 635  -0.928051891      0
#> 636  -0.046118352      0
#> 637   0.214446686      0
#> 638   0.405950457      0
#> 639  -0.340670120      0
#> 640   1.185766673      0
#> 641  -0.177921314      0
#> 642   0.287996667      0
#> 643  -0.213788660      0
#> 644  -0.461231844      0
#> 645  -0.543812943      0
#> 646  -0.110953680      0
#> 647  -0.298225939      0
#> 648  -0.902212335      0
#> 649   0.206464421      0
#> 650   0.202064072      0
#> 651   0.488417048      0
#> 652   0.418592923      0
#> 653  -0.466958838      0
#> 654   0.698109118      0
#> 655   0.030244829      0
#> 656  -0.559357843      0
#> 657  -0.282848814      0
#> 658  -1.087424433      0
#> 659   0.442080093      0
#> 660  -0.717890666      0
#> 661  -0.634175029      0
#> 662  -1.628858706      0
#> 663   1.647963421      0
#> 664  -1.156644474      0
#> 665  -3.142387638      0
#> 666  -0.913892566      0
#> 667   1.306859404      0
#> 668  -0.755536283      0
#> 669  -1.306258549      0
#> 670  -0.060786799      0
#> 671   0.740976209      0
#> 672  -3.445606367      0
#> 673   1.570328366      0
#> 674   0.114115044      0
#> 675   1.245807646      0
#> 676  -0.433390956      0
#> 677  -0.963149327      0
#> 678   1.390945537      0
#> 679   0.278705605      0
#> 680   0.940309872      0
#> 681  -1.446128158      0
#> 682  -0.059325178      0
#> 683   0.110286546      0
#> 684  -0.367874569      0
#> 685   0.274227876      0
#> 686  -0.108508914      0
#> 687   0.414587937      0
#> 688  -0.046275041      0
#> 689  -0.973037806      0
#> 690  -0.488074400      0
#> 691   0.154951835      0
#> 692  -1.322072315      0
#> 693  -1.107182465      0
#> 694  -0.627099307      0
#> 695   0.241038787      0
#> 696   0.757401866      0
#> 697   0.009840924      0
#> 698   0.720797162      0
#> 699  -0.241102486      0
#> 700   1.659213335      0
#> 701  -1.792001796      0
#> 702  -0.045168382      0
#> 703  -0.850728858      0
#> 704  -0.707275596      0
#> 705   1.081694710      0
#> 706   0.236565710      0
#> 707   0.990682463      0
#> 708  -0.212680547      0
#> 709   0.548263607      0
#> 710   1.051892993      0
#> 711  -1.042732879      0
#> 712  -0.161804419      0
#> 713   0.500861754      0
#> 714   1.312061874      0
#> 715   2.522651099      0
#> 716  -1.686225726      0
#> 717  -0.655138937      0
#> 718   1.943527640      0
#> 719  -0.070556894      0
#> 720   1.355463663      0
#> 721   0.818000913      0
#> 722   0.842722494      0
#> 723  -0.659838287      0
#> 724  -0.560019940      0
#> 725  -0.253641324      0
#> 726   2.429362173      0
#> 727   0.902998531      0
#> 728   0.772142591      0
#> 729   0.049715062      0
#> 730  -1.361013886      0
#> 731  -2.489337477      0
#> 732   0.586413188      0
#> 733   0.738566464      0
#> 734  -0.994049909      0
#> 735   0.631428366      0
#> 736   1.111126014      0
#> 737  -1.030074025      0
#> 738   0.520659789      0
#> 739   1.087867253      0
#> 740  -0.176068094      0
#> 741  -0.181605646      0
#> 742  -0.261441372      0
#> 743   0.058053438      0
#> 744  -0.519206220      0
#> 745  -0.836668323      0
#> 746  -0.132444181      0
#> 747  -0.394097259      0
#> 748  -1.513815956      0
#> 749   0.328868362      0
#> 750  -0.183544135      0
#> 751  -1.276579249      0
#> 752  -1.542923843      0
#> 753   0.468859070      0
#> 754  -0.555181705      0
#> 755  -0.490113934      0
#> 756  -0.967431838      0
#> 757   1.706276469      0
#> 758  -0.458991766      0
#> 759   1.165175392      0
#> 760   0.017026728      0
#> 761   0.267984546      0
#> 762   0.163979458      0
#> 763  -1.109309038      0
#> 764   0.285206362      0
#> 765   0.512159074      0
#> 766   0.959368121      0
#> 767  -0.866451000      0
#> 768   0.645928035      0
#> 769   1.185402220      0
#> 770  -1.362750630      0
#> 771  -0.723917375      0
#> 772  -1.343609216      0
#> 773   0.769054614      0
#> 774   1.123052946      0
#> 775   0.831265845      0
#> 776   0.354342486      0
#> 777   0.619975826      0
#> 778   0.106580078      0
#> 779   0.181736168      0
#> 780  -2.142704458      0
#> 781   1.235051252      0
#> 782   0.608050332      0
#> 783   0.330612385      0
#> 784  -1.109446984      0
#> 785   0.642746030      0
#> 786   0.948278355      0
#> 787  -0.001757851      0
#> 788   0.796152292      0
#> 789   2.017115541      0
#> 790   0.543231810      0
#> 791  -1.468935288      0
#> 792  -0.796922002      0
#> 793   1.601629900      0
#> 794   2.208621318      0
#> 795  -1.599459541      0
#> 796   1.778992676      0
#> 797   1.559619573      0
#> 798   2.042863574      0
#> 799   0.993272711      0
#> 800  -0.237455493      0
#> 801   0.149102057      0
#> 802   0.615546082      0
#> 803  -0.285213225      0
#> 804   0.224436606      0
#> 805   1.421146308      0
#> 806  -0.204595179      0
#> 807  -0.844012384      0
#> 808  -2.062470884      0
#> 809   0.903822699      0
#> 810   2.326430860      0
#> 811   3.361844928      0
#> 812   1.769481821      0
#> 813  -1.198406515      0
#> 814  -0.312447204      0
#> 815  -0.044336375      0
#> 816  -0.268641163      0
#> 817  -2.644677674      0
#> 818   0.306160259      0
#> 819   1.235293397      0
#> 820  -0.564739029      0
#> 821  -1.310638692      0
#> 822   0.379882449      0
#> 823   0.200785099      0
#> 824  -0.247463830      0
#> 825   0.302939867      0
#> 826  -0.455669544      0
#> 827   0.587077816      0
#> 828  -1.878301686      0
#> 829  -0.152283729      0
#> 830   1.782970145      0
#> 831  -0.600391781      0
#> 832  -0.360390715      0
#> 833  -0.693231618      0
#> 834  -0.659007749      0
#> 835   1.303364143      0
#> 836   0.469811413      0
#> 837   0.258247976      0
#> 838   1.860433966      0
#> 839  -0.045306084      0
#> 840  -0.902883100      0
#> 841   0.140351475      0
#> 842   0.992332381      0
#> 843  -0.948240480      0
#> 844   0.080826490      0
#> 845  -0.896045884      0
#> 846  -1.370441556      0
#> 847   0.113019005      0
#> 848  -0.523081903      0
#> 849  -0.195780621      0
#> 850  -0.469561206      0
#> 851   0.979257287      0
#> 852  -0.797959445      0
#> 853  -0.397359322      0
#> 854  -1.178900488      0
#> 855   0.631828785      0
#> 856  -1.925076110      0
#> 857  -1.617172726      0
#> 858   0.174299566      0
#> 859  -0.273953776      0
#> 860  -0.079617932      0
#> 861   1.013218319      0
#> 862  -0.060298692      0
#> 863  -1.199405970      0
#> 864  -0.731765592      0
#> 865  -0.385070377      0
#> 866   0.452273511      0
#> 867  -0.289037807      0
#> 868   0.930427241      0
#> 869   0.786414858      0
#> 870  -1.092598581      0
#> 871  -0.962566171      0
#> 872  -0.953495478      0
#> 873   0.790545811      0
#> 874  -0.429026849      0
#> 875   0.197198546      0
#> 876   0.084471860      0
#> 877   0.525728738      0
#> 878   1.029631701      0
#> 879   0.788177838      0
#> 880  -0.915408877      0
#> 881   0.748167140      0
#> 882  -0.191426077      0
#> 883  -0.468922410      0
#> 884  -0.344141251      0
#> 885   1.480031628      0
#> 886  -0.190260887      0
#> 887   0.043361617      0
#> 888   0.913717303      0
#> 889   0.184367481      0
#> 890   0.402349939      0
#> 891  -0.721784067      0
#> 892  -0.334176354      0
#> 893   1.317723827      1
#> 894   1.485895199      1
#> 895   0.536924604      1
#> 896   1.710144757      1
#> 897   0.414869762      1
#> 898   1.704280725      1
#> 899   0.379167502      1
#> 900   0.060368230      1
#> 901   0.017681506      1
#> 902   2.069294870      1
#> 903   2.554752428      1
#> 904   0.343309191      1
#> 905   0.914843251      1
#> 906   1.980656433      1
#> 907   0.502073872      1
#> 908   2.060566565      1
#> 909   1.233090611      1
#> 910   2.229654529      1
#> 911  -0.128081531      1
#> 912   1.692448715      1
#> 913   2.046881380      1
#> 914   2.390309533      1
#> 915   1.091443376      1
#> 916   2.000543938      1
#> 917   0.415427189      1
#> 918   1.312845570      1
#> 919   1.024058657      1
#> 920   0.610113551      1
#> 921  -0.133381459      1
#> 922   0.144953343      1
#> 923  -0.655224007      1
#> 924  -0.271203799      1
#> 925   0.362445359      1
#> 926   0.083283230      1
#> 927   0.313299848      1
#> 928  -0.495846049      1
#> 929  -0.403991565      1
#> 930  -1.069162378      1
#> 931   0.878147699      1
#> 932  -1.006651093      1
#> 933   1.436179790      1
#> 934  -0.330761976      1
#> 935   1.419693103      1
#> 936   1.916026614      1
#> 937  -0.178918399      1
#> 938   0.622184507      1
#> 939   0.013051854      1
#> 940   0.073127568      1
#> 941   1.769538630      1
#> 942   0.253944553      1
#> 943   3.574201989      1
#> 944  -0.006495531      1
#> 945   1.930164961      1
#> 946  -1.229318174      1
#> 947  -0.123086480      1
#> 948   1.001114976      1
#> 949  -0.774296732      1
#> 950   0.142599658      1
#> 951   1.963636531      1
#> 952   0.926116141      1
#> 953   1.192924104      1
#> 954   1.207477638      1
#> 955   0.750407401      1
#> 956   2.607070464      1
#> 957   0.795167613      1
#> 958   0.410186653      1
#> 959   0.845631841      1
#> 960   0.788143896      1
#> 961   0.862626666      1
#> 962   0.961267081      1
#> 963   0.340383383      1
#> 964  -0.110530827      1
#> 965  -0.266313091      1
#> 966  -0.421306447      1
#> 967   0.262982450      1
#> 968  -0.637968402      1
#> 969  -0.120668732      1
#> 970   3.620152391      1
#> 971   0.819392765      1
#> 972   2.906187295      1
#> 973   0.430469556      1
#> 974   0.345287067      1
#> 975   1.040510607      1
#> 976   0.321620874      1
#> 977   0.711237326      1
#> 978  -0.762113435      1
#> 979   1.224001072      1
#> 980  -0.337513777      1
#> 981   1.488943167      1
#> 982   0.575152294      1
#> 983   1.132807759      1
#> 984   0.422291133      1
#> 985   0.111072927      1
#> 986   0.115554871      1
#> 987   0.328324500      1
#> 988  -1.657364157      1
#> 989  -0.314990187      1
#> 990   0.262535400      1
#> 991   1.151683630      1
#> 992   1.165932958      1
#> 993   1.391519442      1
#> 994   0.257516023      1
#> 995   0.277136271      1
#> 996   0.980529748      1
#> 997   2.788041319      1
#> 998  -1.620149377      1
#> 999   0.653895181      1
#> 1000  1.839425300      1