Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     1.2581398520      0
#> 2     0.9181486416      0
#> 3     0.4204801511      0
#> 4     1.6869059564      0
#> 5     1.0227795002      0
#> 6    -0.0151744284      0
#> 7    -0.8056658709      0
#> 8    -0.0870942241      0
#> 9    -0.9090680276      0
#> 10   -0.3234454412      0
#> 11   -0.9356687869      0
#> 12    1.1250387350      0
#> 13   -0.4055557517      0
#> 14   -0.9530020258      0
#> 15    0.4595437163      0
#> 16    0.5702608491      0
#> 17   -0.9510167977      0
#> 18    0.7153299063      0
#> 19    0.0608635139      0
#> 20    1.1582831975      0
#> 21    0.7955581938      0
#> 22   -0.6958613817      0
#> 23   -2.0806340171      0
#> 24    0.9282967975      0
#> 25   -0.0382688811      0
#> 26    0.8902284553      0
#> 27    0.2723233917      0
#> 28   -0.6334490745      0
#> 29   -2.0593972606      0
#> 30    1.0922487975      0
#> 31    0.8958294966      0
#> 32    0.3265889155      0
#> 33   -1.7405242747      0
#> 34    0.0113683761      0
#> 35   -0.3112487915      0
#> 36   -0.2295426207      0
#> 37    1.2194671565      0
#> 38   -0.1363305329      0
#> 39   -0.5962265385      0
#> 40    0.2387579561      0
#> 41    0.0657214452      0
#> 42    0.2194357826      0
#> 43   -0.1666317544      0
#> 44    2.1728013799      0
#> 45    0.7328831280      0
#> 46    0.5206350633      0
#> 47   -1.2447972053      0
#> 48   -0.6280034336      0
#> 49    0.2412883018      0
#> 50   -1.2556668855      0
#> 51    0.7184692611      0
#> 52   -0.5127780975      0
#> 53   -0.0616418412      0
#> 54   -0.7771411859      0
#> 55   -0.3168745638      0
#> 56   -0.1927654906      0
#> 57    0.9744032925      0
#> 58   -0.8639150729      0
#> 59    1.4919543575      0
#> 60   -1.3281877491      0
#> 61   -0.7223699794      0
#> 62   -0.5102040883      0
#> 63   -1.1671024991      0
#> 64    1.3004048771      0
#> 65    0.7214967077      0
#> 66    0.5812858944      0
#> 67   -0.4885853372      0
#> 68    0.3579368580      0
#> 69    1.2459255977      0
#> 70    0.9483300501      0
#> 71   -0.5778015869      0
#> 72   -0.8698055869      0
#> 73   -0.9801661869      0
#> 74   -1.4937689580      0
#> 75    0.2696339268      0
#> 76    1.4425151554      0
#> 77   -1.3594184801      0
#> 78   -0.9279033731      0
#> 79   -0.0509518027      0
#> 80   -0.5174437215      0
#> 81    1.2300902735      0
#> 82   -0.5610997767      0
#> 83   -0.3013635875      0
#> 84   -1.0469072294      0
#> 85    0.6862773657      0
#> 86    0.5822700885      0
#> 87    0.1716483044      0
#> 88   -1.3536492466      0
#> 89   -0.2183056356      0
#> 90    0.5644237644      0
#> 91   -1.1184987579      0
#> 92   -1.7216999963      0
#> 93   -0.6795469409      0
#> 94    1.2473882722      0
#> 95   -1.2070647839      0
#> 96    1.1242332890      0
#> 97    0.5326754895      0
#> 98   -0.5586664279      0
#> 99    0.0392451250      0
#> 100  -0.9133168620      0
#> 101   0.1377536943      0
#> 102  -1.7312172151      0
#> 103   0.8912573300      0
#> 104  -0.0118479298      0
#> 105  -0.4924531509      0
#> 106   0.2994726453      0
#> 107  -0.4031979328      0
#> 108   1.3730796734      0
#> 109   0.2028491773      0
#> 110  -1.4714234602      0
#> 111   0.7601447135      0
#> 112  -0.4031329591      0
#> 113  -0.3536962985      0
#> 114  -0.0987166742      0
#> 115  -1.2050891002      0
#> 116  -0.3545481504      0
#> 117  -1.7546635840      0
#> 118  -0.3487411283      0
#> 119  -0.1179564427      0
#> 120   0.4908319878      0
#> 121  -1.5377883008      0
#> 122   1.0231108860      0
#> 123  -0.3235959235      0
#> 124  -0.3992675421      0
#> 125  -0.1364814342      0
#> 126   0.4645618466      0
#> 127  -2.5456176572      0
#> 128  -0.8186868826      0
#> 129  -0.3493892017      0
#> 130   0.7562460402      0
#> 131   1.0844258817      0
#> 132  -0.7130634924      0
#> 133   0.7965488423      0
#> 134  -0.4608691783      0
#> 135   0.1842580447      0
#> 136   1.0185202116      0
#> 137   0.3353964971      0
#> 138  -1.5210948068      0
#> 139  -1.1234216055      0
#> 140  -1.7130689878      0
#> 141  -0.3763624076      0
#> 142  -0.3317367559      0
#> 143  -0.2356144561      0
#> 144   0.2347029677      0
#> 145   1.1930112255      0
#> 146  -0.0439151245      0
#> 147  -0.3080750079      0
#> 148  -1.3623156789      0
#> 149   1.4311051882      0
#> 150   0.1236283719      0
#> 151   1.0392236745      0
#> 152  -0.8112844726      0
#> 153   0.7283505837      0
#> 154  -0.3801585300      0
#> 155  -1.3355938713      0
#> 156  -0.6894546509      0
#> 157   0.7432655825      0
#> 158   1.0416589822      0
#> 159  -0.5370578827      0
#> 160   0.1251425615      0
#> 161   0.4560764481      0
#> 162   0.9841767902      0
#> 163   0.5539224577      0
#> 164   0.5093928928      0
#> 165  -1.8015884953      0
#> 166   0.5431933362      0
#> 167  -1.0687481059      0
#> 168   0.7339307970      0
#> 169   0.3729668288      0
#> 170  -0.5639789370      0
#> 171  -0.6586523622      0
#> 172  -1.8443578764      0
#> 173  -1.0925824792      0
#> 174   0.4215244195      0
#> 175  -1.4161540667      0
#> 176  -0.2975935985      0
#> 177   0.1518591608      0
#> 178   1.0995452245      0
#> 179  -0.5403049457      0
#> 180   0.0486816875      0
#> 181  -0.9765978317      0
#> 182  -1.0814888877      0
#> 183   1.1426429384      0
#> 184  -0.3640791557      0
#> 185   0.4418482630      0
#> 186   0.3052151687      0
#> 187   0.2102690038      0
#> 188   2.4232362973      0
#> 189   1.3792091726      0
#> 190   0.1345783301      0
#> 191  -1.0718979866      0
#> 192   1.1634048732      0
#> 193   0.1358973791      0
#> 194  -0.4948052789      0
#> 195   0.3051566970      0
#> 196  -0.0241864102      0
#> 197   1.3604444628      0
#> 198  -0.0361937293      0
#> 199   1.4141057838      0
#> 200   0.2338409307      0
#> 201  -0.3492385243      0
#> 202   2.1331776818      0
#> 203   2.1889325054      0
#> 204  -1.6244896110      0
#> 205  -0.6955091071      0
#> 206   0.4187548633      0
#> 207   0.3945306230      0
#> 208  -0.9231627450      0
#> 209   0.8864347800      0
#> 210  -0.1338080260      0
#> 211  -0.9295861546      0
#> 212  -2.0705664371      0
#> 213  -0.9696657065      0
#> 214   0.6197435278      0
#> 215   1.3305630745      0
#> 216   1.8690111285      0
#> 217   0.4491514874      0
#> 218   0.5319946210      0
#> 219  -1.1106774337      0
#> 220  -2.3277826172      0
#> 221  -0.2922171937      0
#> 222   1.7218361068      0
#> 223  -1.1285738013      0
#> 224   0.5755090668      0
#> 225  -1.5236810697      0
#> 226   0.6149664805      0
#> 227  -0.4877207137      0
#> 228  -0.5579717909      0
#> 229  -0.9823151216      0
#> 230  -0.0611633958      0
#> 231  -1.2895699195      0
#> 232  -0.7408429708      0
#> 233  -1.0195537210      0
#> 234   0.7435773594      0
#> 235   0.1605665376      0
#> 236   1.6862914765      0
#> 237   0.6740841199      0
#> 238   0.3544171134      0
#> 239  -1.3937143629      0
#> 240   0.6675680404      0
#> 241  -0.6543290177      0
#> 242  -1.0704758359      0
#> 243   1.8372828781      0
#> 244   1.8702756340      0
#> 245   0.3644872098      0
#> 246  -0.1281235075      0
#> 247   0.7900646620      0
#> 248   0.3199424233      0
#> 249  -0.4574617445      0
#> 250  -1.3346340473      0
#> 251   1.6092175245      0
#> 252   1.0131301725      0
#> 253  -1.6980207731      0
#> 254   0.5660889798      0
#> 255   0.9691214498      0
#> 256  -0.3125950663      0
#> 257  -0.8945994698      0
#> 258   0.1737476028      0
#> 259   0.3339613368      0
#> 260   1.1526365324      0
#> 261   1.4986012439      0
#> 262   0.7136103395      0
#> 263  -0.6935443557      0
#> 264   0.1265018245      0
#> 265   0.6539163585      0
#> 266  -0.6808355436      0
#> 267  -1.0672491334      0
#> 268   0.4749400775      0
#> 269  -1.1333409451      0
#> 270   0.9158792593      0
#> 271   0.0799998975      0
#> 272  -1.6378501675      0
#> 273   0.9617079580      0
#> 274  -1.2800719121      0
#> 275   2.4692928657      0
#> 276  -1.9620564045      0
#> 277   0.6615869906      0
#> 278  -0.9957726824      0
#> 279   1.3137741630      0
#> 280   0.1566410173      0
#> 281   0.3962845364      0
#> 282   1.2121872559      0
#> 283  -0.4915264606      0
#> 284  -0.9675327696      0
#> 285  -1.8184282353      0
#> 286  -0.3242256725      0
#> 287  -0.9681848373      0
#> 288   1.5386947429      0
#> 289   0.5838764629      0
#> 290  -0.5346582665      0
#> 291  -1.0903509709      0
#> 292  -0.7056105960      0
#> 293  -0.7913522000      0
#> 294  -0.4306348650      0
#> 295   0.2730998198      0
#> 296  -1.3622880312      0
#> 297  -0.2388048894      0
#> 298  -0.1808970143      0
#> 299  -0.5392850432      0
#> 300   1.4682654564      0
#> 301   0.7990176618      0
#> 302   1.6175462227      0
#> 303   0.8159508943      0
#> 304  -1.2740181662      0
#> 305   0.0071276862      0
#> 306  -0.1223741883      0
#> 307  -1.3039221164      0
#> 308   1.6295392554      0
#> 309  -0.8678989029      0
#> 310  -0.2459238297      0
#> 311   0.8046116509      0
#> 312   1.8453136584      0
#> 313  -0.7415306316      0
#> 314  -1.4662741457      0
#> 315   0.7631113157      0
#> 316  -0.2965145243      0
#> 317  -0.3109350906      0
#> 318   0.4208620429      0
#> 319  -0.8735436161      0
#> 320   1.4340399626      0
#> 321   1.4643218175      0
#> 322  -0.3281968376      0
#> 323  -0.8549501448      0
#> 324   0.3664954509      0
#> 325   1.9705854859      0
#> 326   1.3352596531      0
#> 327   1.7616420684      0
#> 328   0.6436174898      0
#> 329   1.7589087095      0
#> 330   0.5422146431      0
#> 331   0.3021012979      0
#> 332  -0.9351934543      0
#> 333   0.1899858258      0
#> 334   0.8829784080      0
#> 335  -0.7562364917      0
#> 336  -0.4102562565      0
#> 337   1.5413270682      0
#> 338  -0.0470730679      0
#> 339  -0.7997101289      0
#> 340  -1.0270601276      0
#> 341  -0.5266279584      0
#> 342   0.2422079267      0
#> 343   1.0218117834      0
#> 344  -0.7656912151      0
#> 345   0.4122898477      0
#> 346   1.2179580960      0
#> 347  -0.1643536002      0
#> 348   1.7098878994      0
#> 349   0.3922400393      0
#> 350  -1.3390508262      0
#> 351   0.7549924805      0
#> 352   0.1461666553      0
#> 353   1.4602686173      0
#> 354  -1.8153463694      0
#> 355   0.4390883348      0
#> 356   1.2930395232      0
#> 357  -1.1639565883      0
#> 358   1.1466926811      0
#> 359  -0.5029083550      0
#> 360  -1.3930273166      0
#> 361   0.7789113967      0
#> 362   1.1556438784      0
#> 363  -0.1404219554      0
#> 364  -1.1178619032      0
#> 365  -1.1312166991      0
#> 366   0.4756921006      0
#> 367   0.6024827544      0
#> 368  -0.0400867096      0
#> 369   0.6887837352      0
#> 370  -0.6445592802      0
#> 371  -0.6529581564      0
#> 372  -0.8452844289      0
#> 373   0.6331174960      0
#> 374  -0.0203496906      0
#> 375   0.3685574389      0
#> 376   0.4649920546      0
#> 377   1.4181383135      0
#> 378  -1.1511266749      0
#> 379   0.6672958133      0
#> 380  -0.1916719819      0
#> 381  -0.0184883354      0
#> 382   1.5131342637      0
#> 383   0.1626484089      0
#> 384   1.1195067104      0
#> 385   1.5138986960      0
#> 386  -0.7174093870      0
#> 387   0.2042053467      0
#> 388   0.6939583659      0
#> 389   0.4174182488      0
#> 390  -1.2675820178      0
#> 391   0.2694030851      0
#> 392   0.6267077937      0
#> 393  -1.1277197292      0
#> 394  -0.1864126138      0
#> 395  -0.9527580260      0
#> 396  -0.6560607595      0
#> 397  -1.0249417469      0
#> 398   1.2981488637      0
#> 399   1.3555608086      0
#> 400   0.8831240917      0
#> 401   0.9054507781      0
#> 402  -0.4224886997      0
#> 403   0.8521142607      0
#> 404  -1.5533494519      0
#> 405   0.6021068063      0
#> 406  -1.4947226002      0
#> 407   0.7168291340      0
#> 408  -1.2857872457      0
#> 409  -0.7835354652      0
#> 410  -0.6046472573      0
#> 411  -1.2942623743      0
#> 412   0.7447198653      0
#> 413   0.0458732628      0
#> 414   1.2401785830      0
#> 415   0.2067606104      0
#> 416  -0.6519018743      0
#> 417  -0.4075384945      0
#> 418  -0.2675987350      0
#> 419   0.1396864839      0
#> 420  -0.5354070957      0
#> 421   0.4614112437      0
#> 422  -0.2172008420      0
#> 423  -1.0740174320      0
#> 424   1.1365250558      0
#> 425  -0.2853167871      0
#> 426  -0.5349062908      0
#> 427  -0.0195085883      0
#> 428   2.3559930887      0
#> 429  -0.0109990964      0
#> 430   0.6243457691      0
#> 431   1.1693209412      0
#> 432  -1.1804111604      0
#> 433  -0.6597710298      0
#> 434   0.6008923749      0
#> 435  -0.4903300352      0
#> 436   0.1114020838      0
#> 437   0.7908007187      0
#> 438  -1.3601693062      0
#> 439   1.0211251813      0
#> 440   0.2665797471      0
#> 441  -0.0789628252      0
#> 442   1.7029570320      0
#> 443  -0.5200337213      0
#> 444  -0.4232468483      0
#> 445  -0.8566963261      0
#> 446  -1.9233937992      0
#> 447   0.6569310979      0
#> 448   1.7460237399      0
#> 449  -0.8453345136      0
#> 450   1.7383507747      0
#> 451  -0.7406965093      0
#> 452   1.3651457263      0
#> 453   1.0113435554      0
#> 454  -0.1207916372      0
#> 455   0.5940450362      0
#> 456  -0.3427827164      0
#> 457   0.0312512418      0
#> 458  -0.1546954683      0
#> 459  -1.6435945845      0
#> 460  -0.4279969090      0
#> 461   1.2028762385      0
#> 462  -0.2776444437      0
#> 463  -1.4139531570      0
#> 464   0.4022598523      0
#> 465   0.7422335948      0
#> 466  -1.6386453100      0
#> 467  -0.5948682095      0
#> 468  -1.0937127406      0
#> 469  -2.1556269660      0
#> 470  -0.2441350053      0
#> 471  -0.1883968071      0
#> 472   0.7860457593      0
#> 473  -0.4855167380      0
#> 474   0.3747500848      0
#> 475   0.1313433869      0
#> 476  -0.4373234552      0
#> 477  -1.0526659674      0
#> 478  -0.2869855347      0
#> 479   0.3423153153      0
#> 480  -1.9189467212      0
#> 481  -0.7957635602      0
#> 482  -0.5408646970      0
#> 483   1.7548943761      0
#> 484   0.6578119976      0
#> 485   1.0764401739      0
#> 486   0.5603269249      0
#> 487  -0.1865642970      0
#> 488   0.3217572719      0
#> 489  -1.2317089154      0
#> 490  -1.8695059178      0
#> 491  -0.5230529771      0
#> 492   0.6499330491      0
#> 493  -0.0417095987      0
#> 494  -0.0476597778      0
#> 495   0.5727219921      0
#> 496  -1.8922645725      0
#> 497  -0.3060134919      0
#> 498  -0.3658782271      0
#> 499   0.9477933970      0
#> 500  -0.4332558299      0
#> 501   0.5461281561      0
#> 502  -0.0178164392      0
#> 503   0.0040830238      0
#> 504  -0.3558201063      0
#> 505   0.2668313603      0
#> 506   1.2768088820      0
#> 507   1.6855159099      0
#> 508  -0.7194703777      0
#> 509  -0.6612453671      0
#> 510   1.0220070057      0
#> 511  -1.1228993525      0
#> 512  -0.2195610137      0
#> 513   1.1483293200      0
#> 514  -1.7336838906      0
#> 515   1.6656670059      0
#> 516   1.6377166738      0
#> 517  -0.1785508684      0
#> 518   0.4386375252      0
#> 519   0.8992651404      0
#> 520   2.1547690324      0
#> 521  -0.8197465908      0
#> 522   0.5438273409      0
#> 523   0.6712328626      0
#> 524   0.9140441129      0
#> 525  -0.6375946808      0
#> 526  -0.1387731093      0
#> 527   0.3206728461      0
#> 528   0.0818644769      0
#> 529  -1.8442555028      0
#> 530  -0.4473139698      0
#> 531   0.7290612215      0
#> 532   0.1139887819      0
#> 533   1.3196263000      0
#> 534  -1.3777440747      0
#> 535  -0.6893354799      0
#> 536   0.3265070096      0
#> 537   0.9566041276      0
#> 538   1.1860408574      0
#> 539   1.4090965510      0
#> 540  -1.7013031861      0
#> 541  -0.6761436714      0
#> 542  -0.4946617721      0
#> 543  -1.8940393587      0
#> 544  -0.5549074142      0
#> 545  -0.6039613126      0
#> 546  -0.1128466803      0
#> 547  -0.2771393036      0
#> 548  -1.0766654370      0
#> 549   0.5244776332      0
#> 550  -1.0700325021      0
#> 551  -0.6072179511      0
#> 552  -0.9496452309      0
#> 553  -0.5941130297      0
#> 554  -1.0765800643      0
#> 555   0.0680091524      0
#> 556  -0.6234932151      0
#> 557  -0.6471414416      0
#> 558   0.1261770999      0
#> 559  -1.3164984020      0
#> 560  -0.3863907116      0
#> 561  -0.0831514204      0
#> 562   0.8946604093      0
#> 563   1.1065608925      0
#> 564   0.1972945378      0
#> 565  -0.4589774129      0
#> 566  -0.2611343204      0
#> 567   0.3993697907      0
#> 568   0.5828586051      0
#> 569  -1.1428396948      0
#> 570   0.7547991200      0
#> 571   0.0455494423      0
#> 572   1.7376698915      0
#> 573   0.5481614226      0
#> 574  -0.7627521140      0
#> 575   0.3369158187      0
#> 576   0.6242552940      0
#> 577   0.6514216520      0
#> 578  -0.0113667636      0
#> 579   1.0801846204      0
#> 580  -0.0081656398      0
#> 581  -0.7396017771      0
#> 582  -1.6156030354      0
#> 583   2.8611607326      0
#> 584  -1.8881897667      0
#> 585   0.8791368557      0
#> 586  -0.4060924869      0
#> 587   0.2513959991      0
#> 588  -1.4409829175      0
#> 589  -0.3200032326      0
#> 590  -2.9810004792      0
#> 591   0.6077429306      0
#> 592   0.6072384433      0
#> 593   0.3151930287      0
#> 594   3.5995227451      0
#> 595  -0.8488696453      0
#> 596   0.6620441550      0
#> 597   1.4512871016      0
#> 598  -0.0776777806      0
#> 599   1.1744339812      0
#> 600  -0.5610822612      0
#> 601   0.5368366594      0
#> 602   0.9214493798      0
#> 603  -1.5412658084      0
#> 604  -1.8134075106      0
#> 605  -0.9089027354      0
#> 606   0.7264951065      0
#> 607   0.2570748150      0
#> 608  -0.0521114682      0
#> 609  -1.6366167456      0
#> 610  -1.0676222847      0
#> 611   1.7648069889      0
#> 612  -1.1632871363      0
#> 613  -0.2484497038      0
#> 614   0.6587581118      0
#> 615  -2.4738552336      0
#> 616   1.1671023805      0
#> 617  -1.7156492207      0
#> 618  -1.0878411378      0
#> 619   0.7938985796      0
#> 620   0.0990095907      0
#> 621  -1.5767483879      0
#> 622  -0.3520752684      0
#> 623  -0.5414241491      0
#> 624   0.4287682629      0
#> 625  -1.2469063789      0
#> 626  -0.2605069707      0
#> 627   0.0339807848      0
#> 628  -0.1934767991      0
#> 629   0.2083842834      0
#> 630   0.2261621670      0
#> 631  -0.7753278687      0
#> 632   1.8607383184      0
#> 633  -1.1191251794      0
#> 634   1.4466123572      0
#> 635   0.1866218092      0
#> 636  -0.9120522553      0
#> 637   0.2273349504      0
#> 638  -0.6661395387      0
#> 639   0.4935884405      0
#> 640  -0.0850126610      0
#> 641   1.0050587893      0
#> 642  -0.2968573878      0
#> 643  -1.0904378099      0
#> 644  -0.9767995172      0
#> 645  -0.1735801546      0
#> 646  -0.0269389417      0
#> 647   1.0156356958      0
#> 648  -0.2586509860      0
#> 649  -0.6760372826      0
#> 650  -0.0219855412      0
#> 651   0.8621309938      0
#> 652  -0.1616306829      0
#> 653  -2.3862632382      0
#> 654   0.2627010173      0
#> 655  -1.0966072909      0
#> 656  -0.4964986735      0
#> 657  -0.4217276169      0
#> 658  -1.4199788446      0
#> 659  -0.7398384402      0
#> 660  -0.4336603133      0
#> 661   1.7388352141      0
#> 662  -0.5638951054      0
#> 663   0.2026076395      0
#> 664   0.6495633864      0
#> 665   0.8378884705      0
#> 666   1.1034393588      0
#> 667  -1.4862915827      0
#> 668   0.9238525693      0
#> 669  -0.7007027769      0
#> 670  -0.4690454930      0
#> 671  -0.8684363739      0
#> 672   0.4886929769      0
#> 673   0.3730229680      0
#> 674  -0.0220582875      0
#> 675  -0.6393277212      0
#> 676   1.9741910273      0
#> 677  -0.1874729064      0
#> 678   1.1887352661      0
#> 679   2.0744403733      0
#> 680  -0.2175361144      0
#> 681   0.9396025942      0
#> 682  -0.0740898128      0
#> 683  -0.9750777796      0
#> 684   1.2962791663      0
#> 685  -0.2197681617      0
#> 686  -0.0358819881      0
#> 687  -0.8406883763      0
#> 688   1.0771917258      0
#> 689   0.0010593737      0
#> 690  -0.4489225434      0
#> 691  -1.3386105996      0
#> 692  -0.0945116832      0
#> 693   1.4520550453      0
#> 694   0.8876017837      0
#> 695  -0.7444220447      0
#> 696   0.5988966664      0
#> 697  -1.6439374753      0
#> 698   0.8029052657      0
#> 699  -1.0891656395      0
#> 700   0.4678323641      0
#> 701  -1.9391747988      0
#> 702   2.4310857312      0
#> 703   0.4210531000      0
#> 704  -0.0806828061      0
#> 705  -1.1689621909      0
#> 706   1.1831836138      0
#> 707  -0.3062431440      0
#> 708  -0.1715257660      0
#> 709   0.5372997426      0
#> 710  -1.3456110104      0
#> 711   1.1640566305      0
#> 712   0.8745496777      0
#> 713  -1.1654680295      0
#> 714   0.3092639537      0
#> 715  -0.3130910419      0
#> 716   1.5854701697      0
#> 717   0.4063025287      0
#> 718   0.4898034591      0
#> 719   1.2521396630      0
#> 720   0.1746015821      0
#> 721  -1.2857711439      0
#> 722  -0.4910939394      0
#> 723   1.2901246305      0
#> 724   0.4269176036      0
#> 725   1.6821514332      0
#> 726  -0.0302544692      0
#> 727  -0.9818622842      0
#> 728  -1.9008781852      0
#> 729  -0.0860157107      0
#> 730   1.4485377758      0
#> 731   1.2412270911      0
#> 732  -0.2372894211      0
#> 733   0.3484808131      0
#> 734  -1.2583819407      0
#> 735   0.4777771729      0
#> 736  -0.5954176858      0
#> 737   0.0008844661      0
#> 738  -0.7190673354      0
#> 739  -0.5998050507      0
#> 740  -0.2153514017      0
#> 741   2.4142897971      0
#> 742   1.1352450898      0
#> 743   1.1594048930      0
#> 744  -0.1110121759      0
#> 745  -0.6719935586      0
#> 746   0.3050538375      0
#> 747   0.4535659014      0
#> 748  -1.1711105776      0
#> 749  -0.3506963857      0
#> 750  -0.0061883589      0
#> 751   0.5677853501      0
#> 752  -1.5013989103      0
#> 753   1.0168765148      0
#> 754   0.3378411083      0
#> 755  -1.4643978787      0
#> 756  -0.1335196031      0
#> 757   0.8868374378      0
#> 758   1.3171249499      0
#> 759   2.0911918033      0
#> 760  -0.0313298029      0
#> 761   0.3738231021      0
#> 762  -1.1608364851      0
#> 763  -0.9958388670      0
#> 764   1.0042480556      0
#> 765   0.2907386648      0
#> 766   0.9979259967      0
#> 767  -0.1455002196      0
#> 768  -0.2719682989      0
#> 769   0.9364656775      0
#> 770   1.0407716801      0
#> 771   0.9058099259      0
#> 772   0.0603238697      0
#> 773   0.0009241497      0
#> 774  -0.0652563576      0
#> 775  -0.9428326552      0
#> 776  -0.7644904981      0
#> 777  -1.4242618242      0
#> 778  -0.5755442487      0
#> 779  -0.1597275597      0
#> 780   1.0169207764      0
#> 781   1.1380278736      0
#> 782   0.3135390348      0
#> 783   0.7083432132      0
#> 784   1.0709720564      0
#> 785   0.0137470163      0
#> 786   0.3819217124      0
#> 787  -0.8347552718      0
#> 788   0.3824840371      0
#> 789   0.6836328270      0
#> 790   0.5828763315      0
#> 791  -0.6980974151      0
#> 792  -0.1635616224      0
#> 793   0.7109817193      0
#> 794   0.3966978243      0
#> 795   0.8816070483      0
#> 796   0.6974318545      0
#> 797  -0.1334884391      0
#> 798   0.2520910531      0
#> 799  -0.2234461413      0
#> 800   0.0689765065      0
#> 801  -0.5473949756      0
#> 802  -0.1002798532      0
#> 803   0.2962353978      0
#> 804  -0.3704852905      0
#> 805  -1.5228078367      0
#> 806   0.3927654011      0
#> 807   0.4551792637      0
#> 808   0.0048747495      0
#> 809   0.6796728723      0
#> 810   0.3909415913      0
#> 811   0.3916151253      0
#> 812   0.3362952344      0
#> 813   1.9601011453      0
#> 814   0.6435018501      0
#> 815   0.5943401105      0
#> 816  -0.7974499149      0
#> 817   0.3349805888      0
#> 818  -1.7306566054      0
#> 819  -0.5829531092      0
#> 820  -0.9286984796      0
#> 821  -1.0959968731      0
#> 822   1.2636004775      0
#> 823  -0.1459564202      0
#> 824   1.2489234047      0
#> 825  -0.6686034015      0
#> 826  -1.2608190662      0
#> 827   0.3839446265      0
#> 828  -1.7326286124      0
#> 829   0.0185212728      0
#> 830   0.1509033931      0
#> 831   0.2827198502      0
#> 832   0.5775438051      0
#> 833   0.0538774444      0
#> 834  -0.3660485991      0
#> 835  -1.1135865043      0
#> 836  -1.6439865697      0
#> 837  -0.8683957101      0
#> 838  -1.1878050280      0
#> 839  -0.3529978739      0
#> 840  -0.9090290544      0
#> 841   0.7967093111      0
#> 842  -1.0487134935      0
#> 843  -0.2131510747      0
#> 844   0.5015766284      0
#> 845   0.0970052026      0
#> 846   0.0856257852      0
#> 847   0.4465865973      0
#> 848   0.3736858098      0
#> 849   0.8418991711      0
#> 850   3.0042404309      0
#> 851  -0.3892018134      0
#> 852   1.3673904862      0
#> 853  -1.5672795800      0
#> 854  -1.5829393424      0
#> 855   0.0884862463      0
#> 856  -0.9853982375      0
#> 857  -1.3073506821      0
#> 858   0.5110207252      0
#> 859  -1.2304644042      0
#> 860  -1.3280165039      0
#> 861   0.1626695394      0
#> 862  -0.8204857523      0
#> 863   0.2470211291      0
#> 864  -0.3641230457      0
#> 865  -0.2191847860      0
#> 866  -0.7484281071      0
#> 867   0.1300580091      0
#> 868  -2.3316360953      0
#> 869   0.9547159778      0
#> 870   1.1083184757      0
#> 871  -1.9234597283      0
#> 872   0.0152829412      0
#> 873  -1.0790603825      0
#> 874  -0.9461204066      0
#> 875  -0.5577276339      0
#> 876   1.9333601871      0
#> 877   0.5027336180      0
#> 878   0.4672961453      0
#> 879   0.6352386267      0
#> 880   0.9314272388      0
#> 881  -0.7833477595      0
#> 882   1.3035465547      0
#> 883   1.4202820362      0
#> 884  -0.4547337285      0
#> 885  -1.2988299530      0
#> 886  -1.4483460791      0
#> 887  -0.4668234152      0
#> 888   0.6274988588      0
#> 889  -1.2086764309      0
#> 890   0.7286584780      0
#> 891  -1.0539710149      0
#> 892  -0.6727920257      0
#> 893   0.9453161131      0
#> 894  -0.8832086217      0
#> 895   0.1009415131      0
#> 896  -0.6768990723      0
#> 897   0.4233964375      0
#> 898  -1.7836526405      0
#> 899   0.8040715766      0
#> 900  -0.2011750096      0
#> 901   2.0684186508      1
#> 902   0.3670217765      1
#> 903  -0.0233815022      1
#> 904  -0.0374845995      1
#> 905   1.4799572940      1
#> 906   0.4388814515      1
#> 907   0.8030968464      1
#> 908  -0.6274034213      1
#> 909   1.2353822089      1
#> 910   1.0680255794      1
#> 911   0.7507659733      1
#> 912  -0.0933516534      1
#> 913   0.8647795622      1
#> 914   1.3897347640      1
#> 915   0.7737879494      1
#> 916   2.1926496135      1
#> 917   0.0237494780      1
#> 918   1.2920443430      1
#> 919   0.6754588591      1
#> 920   2.2228184682      1
#> 921  -0.0643506220      1
#> 922   1.3431527749      1
#> 923  -1.4829914243      1
#> 924  -0.1424675263      1
#> 925   0.3548179715      1
#> 926   0.8103338651      1
#> 927   1.8090886850      1
#> 928   2.4200373333      1
#> 929   0.4147445685      1
#> 930   1.8681620489      1
#> 931   0.2135697076      1
#> 932   1.3829812906      1
#> 933   1.4090902071      1
#> 934  -1.1996795774      1
#> 935   2.7287248897      1
#> 936  -0.0325127849      1
#> 937  -0.9348435467      1
#> 938   0.4931531954      1
#> 939   1.1859378686      1
#> 940   0.1936897740      1
#> 941  -0.3844546400      1
#> 942  -0.7819973429      1
#> 943   1.2625400021      1
#> 944  -0.3035181199      1
#> 945   1.5735832147      1
#> 946   1.5091159630      1
#> 947   0.5952816225      1
#> 948   0.4990536863      1
#> 949   0.3245111660      1
#> 950   0.8193951997      1
#> 951   0.6832858030      1
#> 952   0.8678341430      1
#> 953   1.2019159941      1
#> 954   1.8336243787      1
#> 955   0.4842099275      1
#> 956   1.9500348518      1
#> 957   1.4630909445      1
#> 958   1.0289577930      1
#> 959  -0.8895155903      1
#> 960   0.4922012989      1
#> 961  -0.4565839338      1
#> 962   2.2242843713      1
#> 963  -0.1082863041      1
#> 964   0.4018927323      1
#> 965   0.8331216764      1
#> 966   0.3653630233      1
#> 967  -0.4890540885      1
#> 968  -0.7289892497      1
#> 969  -0.0521996923      1
#> 970   0.0939696234      1
#> 971   0.1216856603      1
#> 972  -1.5051473132      1
#> 973   0.0951905264      1
#> 974  -0.1821618454      1
#> 975   1.5229515522      1
#> 976   1.8353147308      1
#> 977   1.4086752635      1
#> 978  -0.7522331957      1
#> 979   1.2936225821      1
#> 980  -0.6040535974      1
#> 981   1.7638306710      1
#> 982   0.0313621599      1
#> 983   2.1315336794      1
#> 984   0.9519740940      1
#> 985  -0.2312417066      1
#> 986  -0.0247164719      1
#> 987   0.8944264329      1
#> 988   1.1203049887      1
#> 989  -0.0778963513      1
#> 990   1.7638638278      1
#> 991   1.1024150505      1
#> 992   0.1636561062      1
#> 993  -0.0309652249      1
#> 994   0.1755149778      1
#> 995   1.5161750389      1
#> 996  -0.2512344245      1
#> 997  -0.4939471987      1
#> 998   0.1882561739      1
#> 999   1.2334850951      1
#> 1000  2.3788753537      1