Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     1.733586620      0
#> 2    -1.776907344      0
#> 3    -1.727213003      0
#> 4    -0.414083244      0
#> 5    -0.042765891      0
#> 6    -1.049121820      0
#> 7    -1.219320861      0
#> 8    -1.957104716      0
#> 9    -0.974479319      0
#> 10    0.419504331      0
#> 11   -0.487422476      0
#> 12    0.552183709      0
#> 13   -0.012209982      0
#> 14   -0.310662821      0
#> 15   -0.129750273      0
#> 16    0.458272498      0
#> 17    2.655249933      0
#> 18    1.855060780      0
#> 19   -0.772273883      0
#> 20   -2.276837551      0
#> 21    0.522923392      0
#> 22   -2.685081739      0
#> 23    0.963221370      0
#> 24   -0.101650697      0
#> 25    1.545089673      0
#> 26   -0.981207181      0
#> 27    0.395892151      0
#> 28    0.362564983      0
#> 29   -0.907685309      0
#> 30   -0.288598309      0
#> 31   -1.227220968      0
#> 32   -1.612975543      0
#> 33   -0.591363266      0
#> 34   -0.248382814      0
#> 35    0.628455857      0
#> 36    1.546829075      0
#> 37   -1.571858308      0
#> 38   -0.207690892      0
#> 39    1.712870090      0
#> 40    0.022840317      0
#> 41   -0.523870432      0
#> 42    0.192777824      0
#> 43    0.412186549      0
#> 44   -0.593638053      0
#> 45   -0.804500491      0
#> 46    0.205379507      0
#> 47    1.972654303      0
#> 48   -0.550192589      0
#> 49   -1.092994942      0
#> 50    0.851885550      0
#> 51    0.753867385      0
#> 52   -0.645612445      0
#> 53    1.222878495      0
#> 54    0.360794334      0
#> 55   -0.142119662      0
#> 56    0.429460428      0
#> 57   -0.278906299      0
#> 58   -1.476392886      0
#> 59    0.122951959      0
#> 60    1.278912928      0
#> 61   -1.379047057      0
#> 62    0.883739981      0
#> 63    1.421843833      0
#> 64    1.391586704      0
#> 65    0.064652166      0
#> 66    0.560979363      0
#> 67    0.177329068      0
#> 68   -1.771847949      0
#> 69   -0.630331928      0
#> 70   -0.368073017      0
#> 71   -0.066363911      0
#> 72    1.370179805      0
#> 73   -0.709852947      0
#> 74    0.063280746      0
#> 75   -1.912490278      0
#> 76   -1.910933730      0
#> 77   -1.551804597      0
#> 78    0.564461668      0
#> 79    0.767572321      0
#> 80   -0.882589752      0
#> 81   -0.062169561      0
#> 82   -0.880399726      0
#> 83   -0.106071524      0
#> 84    0.306117224      0
#> 85    0.332212179      0
#> 86    0.092319762      0
#> 87   -1.216368103      0
#> 88   -0.636454317      0
#> 89    0.035371569      0
#> 90    0.720335854      0
#> 91    1.004571842      0
#> 92    1.016487188      0
#> 93   -0.696284154      0
#> 94    0.233780432      0
#> 95   -0.107008577      0
#> 96    1.412173305      0
#> 97    0.620071929      0
#> 98    0.671320646      0
#> 99    0.797265997      0
#> 100  -0.259537376      0
#> 101  -0.227576010      0
#> 102   2.478737438      0
#> 103   1.187663240      0
#> 104  -0.313273501      0
#> 105   0.906709776      0
#> 106  -0.317600210      0
#> 107   0.663870548      0
#> 108  -0.608331033      0
#> 109   1.659657451      0
#> 110  -0.937198352      0
#> 111   1.374893771      0
#> 112  -0.730444838      0
#> 113  -0.308049671      0
#> 114  -2.028279899      0
#> 115   0.947687272      0
#> 116   0.183120916      0
#> 117   0.196467489      0
#> 118  -2.502211892      0
#> 119   0.628294368      0
#> 120  -0.497682122      0
#> 121   0.277432105      0
#> 122   1.612019432      0
#> 123   0.171726121      0
#> 124  -0.226042514      0
#> 125  -1.306176937      0
#> 126  -0.445439188      0
#> 127   0.301851146      0
#> 128   1.100954104      0
#> 129  -2.047062560      0
#> 130  -0.292418010      0
#> 131  -1.461809656      0
#> 132   0.665143420      0
#> 133  -1.440871602      0
#> 134  -1.265574613      0
#> 135  -1.216923261      0
#> 136   0.822995184      0
#> 137   0.540579706      0
#> 138  -1.608616620      0
#> 139  -0.366963946      0
#> 140  -1.015754709      0
#> 141   1.123714950      0
#> 142   1.358229168      0
#> 143   1.515287879      0
#> 144   1.018968792      0
#> 145  -1.633630398      0
#> 146  -0.180125833      0
#> 147   0.197272255      0
#> 148   0.546321142      0
#> 149   0.111396003      0
#> 150   0.629842188      0
#> 151  -0.459579032      0
#> 152   0.997453715      0
#> 153   0.224863762      0
#> 154   0.107513249      0
#> 155  -0.086265449      0
#> 156  -0.149161025      0
#> 157  -2.114448224      0
#> 158   0.047956252      0
#> 159  -0.482596279      0
#> 160   1.014065813      0
#> 161   0.254014037      0
#> 162   0.168010530      0
#> 163   0.948841525      0
#> 164   0.424241310      0
#> 165   0.469053455      0
#> 166   0.164703307      0
#> 167   0.341098290      0
#> 168  -0.193668397      0
#> 169  -1.840948561      0
#> 170  -1.152763564      0
#> 171   1.876665510      0
#> 172  -0.282601243      0
#> 173  -1.118855014      0
#> 174  -1.632605982      0
#> 175   0.894157513      0
#> 176  -1.077079192      0
#> 177   1.146121801      0
#> 178  -0.152947617      0
#> 179  -0.395312093      0
#> 180   0.469175019      0
#> 181   0.774856392      0
#> 182  -0.579053858      0
#> 183  -0.968594795      0
#> 184  -1.491340131      0
#> 185  -0.405972779      0
#> 186   1.566216595      0
#> 187   2.096201952      0
#> 188   0.984748300      0
#> 189   0.210481781      0
#> 190  -0.718611655      0
#> 191   0.171553204      0
#> 192   0.854420380      0
#> 193  -0.511467961      0
#> 194  -0.746761944      0
#> 195   0.469176392      0
#> 196  -0.422482365      0
#> 197   1.386816469      0
#> 198  -0.325268914      0
#> 199   0.855231301      0
#> 200   0.825046608      0
#> 201  -1.051901469      0
#> 202   0.119418613      0
#> 203   2.459064692      0
#> 204  -0.090747315      0
#> 205   1.432953523      0
#> 206  -1.405770518      0
#> 207  -1.055069385      0
#> 208  -1.959980075      0
#> 209   0.693231262      0
#> 210  -2.912018013      0
#> 211   2.391895929      0
#> 212  -0.625763924      0
#> 213  -0.328688519      0
#> 214   0.623063411      0
#> 215  -0.387235672      0
#> 216   1.600515176      0
#> 217   0.664854801      0
#> 218  -0.035069386      0
#> 219  -0.794016971      0
#> 220   0.531775435      0
#> 221   2.145274648      0
#> 222  -0.951173523      0
#> 223   0.451019558      0
#> 224   0.742630635      0
#> 225  -1.060939713      0
#> 226  -1.944151575      0
#> 227   0.091275738      0
#> 228  -0.200985927      0
#> 229  -0.813158629      0
#> 230  -0.809628666      0
#> 231  -1.951479540      0
#> 232   1.403438515      0
#> 233  -0.572605909      0
#> 234   1.066259952      0
#> 235  -1.681287568      0
#> 236  -0.063796421      0
#> 237   0.339386754      0
#> 238   1.121392837      0
#> 239  -1.738226473      0
#> 240   0.985223525      0
#> 241  -0.133133804      0
#> 242  -1.220560453      0
#> 243  -0.027225936      0
#> 244  -0.047169276      0
#> 245  -0.850569796      0
#> 246   0.538269494      0
#> 247  -0.191406718      0
#> 248   0.176444477      0
#> 249   0.924879745      0
#> 250   0.912941529      0
#> 251   0.525372605      0
#> 252   0.752703541      0
#> 253   1.826507348      0
#> 254   2.450608634      0
#> 255   0.591070833      0
#> 256  -0.758445575      0
#> 257   1.389683047      0
#> 258  -0.240348090      0
#> 259   0.911189571      0
#> 260  -0.282209575      0
#> 261   0.291249704      0
#> 262  -0.113112025      0
#> 263  -0.739056819      0
#> 264   1.077029275      0
#> 265  -1.095990051      0
#> 266  -0.554925965      0
#> 267  -0.514436072      0
#> 268   1.148144968      0
#> 269  -0.250395922      0
#> 270   0.169018497      0
#> 271  -0.359571318      0
#> 272   0.257534312      0
#> 273  -1.115460058      0
#> 274   0.484094513      0
#> 275   2.486263852      0
#> 276  -0.382278643      0
#> 277   1.406764056      0
#> 278  -0.671702901      0
#> 279   0.702745137      0
#> 280   1.027559132      0
#> 281  -0.928349534      0
#> 282  -1.122686031      0
#> 283   0.027631466      0
#> 284   0.384788611      0
#> 285   0.312181820      0
#> 286  -0.968275613      0
#> 287   0.620815357      0
#> 288   0.212850313      0
#> 289  -0.136625258      0
#> 290  -0.400937840      0
#> 291   0.551632496      0
#> 292  -0.347829405      0
#> 293  -0.334966766      0
#> 294  -1.935466150      0
#> 295  -0.434785226      0
#> 296   0.333417211      0
#> 297  -0.894519473      0
#> 298   0.314722397      0
#> 299   0.610652425      0
#> 300  -1.250314751      0
#> 301  -0.328311308      0
#> 302   0.249952399      0
#> 303   0.726375463      0
#> 304  -0.143248913      0
#> 305  -0.423808583      0
#> 306  -2.323082854      0
#> 307   2.164880538      0
#> 308  -0.837910031      0
#> 309  -0.940046310      0
#> 310  -0.034384795      0
#> 311   1.392730493      0
#> 312   1.357490723      0
#> 313   1.108419858      0
#> 314  -0.995564163      0
#> 315   0.424117602      0
#> 316   0.266459050      0
#> 317   1.384740729      0
#> 318  -0.896864478      0
#> 319  -0.875502766      0
#> 320   0.245316043      0
#> 321  -0.436624092      0
#> 322  -0.005488141      0
#> 323  -0.485544683      0
#> 324  -1.009583056      0
#> 325   0.665284483      0
#> 326  -0.094221263      0
#> 327   1.293982932      0
#> 328   0.939891673      0
#> 329  -0.933206840      0
#> 330  -1.549604683      0
#> 331   0.250928224      0
#> 332   0.603063710      0
#> 333  -0.894972559      0
#> 334  -0.243411775      0
#> 335  -0.597040881      0
#> 336   0.261071562      0
#> 337  -1.271949282      0
#> 338  -1.259243449      0
#> 339  -0.888535068      0
#> 340   0.032259688      0
#> 341   0.613722729      0
#> 342  -0.264441467      0
#> 343  -0.354605174      0
#> 344   0.462854128      0
#> 345   1.181029092      0
#> 346  -1.911516727      0
#> 347   0.392333859      0
#> 348  -0.918665786      0
#> 349   1.072626410      0
#> 350  -0.873749223      0
#> 351   1.592892274      0
#> 352   0.563408579      0
#> 353  -0.533174709      0
#> 354   0.065579644      0
#> 355  -1.212600148      0
#> 356  -1.115019577      0
#> 357  -1.839035044      0
#> 358  -0.885022137      0
#> 359  -0.110076125      0
#> 360   0.711588233      0
#> 361  -0.516637886      0
#> 362  -0.677487768      0
#> 363  -0.694830630      0
#> 364   1.385343237      0
#> 365  -1.495719928      0
#> 366   0.677979826      0
#> 367  -0.274017972      0
#> 368   0.482659072      0
#> 369  -0.027122055      0
#> 370   1.579234446      0
#> 371  -0.930410336      0
#> 372   1.852859807      0
#> 373   1.850016413      0
#> 374  -1.725143169      0
#> 375  -0.846103741      0
#> 376  -1.087134113      0
#> 377   0.930162041      0
#> 378  -0.352388181      0
#> 379  -0.345834221      0
#> 380  -0.120630812      0
#> 381  -0.036441640      0
#> 382   0.475503138      0
#> 383  -0.597586894      0
#> 384   0.318092970      0
#> 385  -1.311350411      0
#> 386  -0.524685501      0
#> 387  -0.692258939      0
#> 388   1.701176933      0
#> 389   0.591588022      0
#> 390  -0.417962703      0
#> 391   1.240294397      0
#> 392  -0.825707909      0
#> 393   1.573948781      0
#> 394  -0.389617287      0
#> 395  -0.683579529      0
#> 396   1.069207925      0
#> 397   0.127673623      0
#> 398   0.139311955      0
#> 399   0.358595453      0
#> 400  -0.104365333      0
#> 401   0.712914772      0
#> 402   1.990682317      0
#> 403   0.765085636      0
#> 404  -0.312400309      0
#> 405   2.003750144      0
#> 406   1.292967282      0
#> 407  -1.611937628      0
#> 408  -0.768599818      0
#> 409  -0.624700727      0
#> 410  -1.179372220      0
#> 411  -1.575630085      0
#> 412  -1.326363484      0
#> 413  -0.306721442      0
#> 414  -2.062395482      0
#> 415  -0.515382017      0
#> 416   0.964374601      0
#> 417  -1.400850299      0
#> 418  -2.852688156      0
#> 419   0.667989969      0
#> 420  -0.659935312      0
#> 421   2.119490599      0
#> 422  -0.660997749      0
#> 423   1.849796370      0
#> 424   0.836417821      0
#> 425  -1.260352584      0
#> 426   0.654754059      0
#> 427  -0.006125992      0
#> 428   1.191560542      0
#> 429   0.108064166      0
#> 430   1.958155399      0
#> 431  -1.155483599      0
#> 432  -0.583025530      0
#> 433   0.637548246      0
#> 434   0.648681271      0
#> 435  -0.548195882      0
#> 436   0.834685416      0
#> 437  -1.144809719      0
#> 438   0.999269183      0
#> 439   1.515336547      0
#> 440  -0.650917812      0
#> 441   1.348833379      0
#> 442  -1.202882968      0
#> 443   0.374833267      0
#> 444  -0.746980185      0
#> 445   0.058064645      0
#> 446  -0.996739292      0
#> 447  -0.076841964      0
#> 448   0.058962799      0
#> 449  -3.407534965      0
#> 450   0.159561889      0
#> 451   0.899503840      0
#> 452   0.132598038      0
#> 453  -1.264779549      0
#> 454   0.604204668      0
#> 455   2.332885497      0
#> 456  -1.857540067      0
#> 457   0.680088482      0
#> 458   0.114875682      0
#> 459   1.494268931      0
#> 460   1.626689427      0
#> 461  -0.013323845      0
#> 462   1.146664005      0
#> 463  -0.322391719      0
#> 464   0.436102155      0
#> 465  -0.366800717      0
#> 466  -0.315641860      0
#> 467  -0.021024861      0
#> 468  -0.836979402      0
#> 469   0.726603009      0
#> 470  -1.712249261      0
#> 471  -0.016557657      0
#> 472  -0.663262633      0
#> 473   1.103256593      0
#> 474  -3.116276081      0
#> 475   1.558358947      0
#> 476  -0.217432103      0
#> 477   0.043584458      0
#> 478   0.502788297      0
#> 479  -0.004893212      0
#> 480  -0.361728330      0
#> 481   0.617424809      0
#> 482   0.010064168      0
#> 483   1.556008691      0
#> 484   0.610303917      0
#> 485   0.734748444      0
#> 486  -0.013343920      0
#> 487   0.885025701      0
#> 488  -0.211507365      0
#> 489  -0.316699357      0
#> 490   0.328816253      0
#> 491  -0.561144652      0
#> 492  -1.103057411      0
#> 493   1.298938608      0
#> 494  -0.536839779      0
#> 495   2.006232848      0
#> 496   1.341571049      0
#> 497  -0.286861223      0
#> 498   0.543791996      0
#> 499   0.977613788      0
#> 500   0.594057015      0
#> 501   0.264543913      0
#> 502  -1.089652702      0
#> 503   0.082105174      0
#> 504  -0.433134051      0
#> 505   0.893016824      0
#> 506   1.181942821      0
#> 507   1.215489910      0
#> 508  -0.241057564      0
#> 509   0.689989715      0
#> 510  -0.179586552      0
#> 511   0.330007119      0
#> 512   0.174279175      0
#> 513   0.028725804      0
#> 514  -1.525424017      0
#> 515   1.038936896      0
#> 516  -0.217623212      0
#> 517  -1.181160572      0
#> 518   0.640995263      0
#> 519  -1.248173537      0
#> 520   0.026134989      0
#> 521  -0.517416722      0
#> 522   0.555011186      0
#> 523  -0.281263515      0
#> 524   0.741018918      0
#> 525   0.276782309      0
#> 526  -0.079681471      0
#> 527   1.184537286      0
#> 528  -2.033982901      0
#> 529   1.152459215      0
#> 530   1.189683101      0
#> 531   0.205648511      0
#> 532  -0.223787579      0
#> 533  -0.288599445      0
#> 534   1.037982005      0
#> 535   0.200369063      0
#> 536   2.472275497      0
#> 537   1.301891338      0
#> 538   1.138785669      0
#> 539  -1.140202643      0
#> 540   0.231268235      0
#> 541   2.563435554      0
#> 542   0.171605071      0
#> 543   0.761497636      0
#> 544   0.927489927      0
#> 545   0.016071370      0
#> 546  -0.155115439      0
#> 547  -0.792104491      0
#> 548   1.849318984      0
#> 549   0.644985347      0
#> 550  -1.032830770      0
#> 551  -0.152441372      0
#> 552  -0.205514155      0
#> 553   0.200859499      0
#> 554   0.071372918      0
#> 555   1.062477637      0
#> 556  -0.199334672      0
#> 557   0.516440941      0
#> 558  -0.293390332      0
#> 559   0.130388246      0
#> 560  -0.719003248      0
#> 561  -1.610858907      0
#> 562  -1.121746102      0
#> 563   0.109324949      0
#> 564  -0.579752260      0
#> 565  -1.199386850      0
#> 566  -1.404521556      0
#> 567  -0.532580515      0
#> 568   0.024151034      0
#> 569   0.833858212      0
#> 570  -0.596782787      0
#> 571   1.627501059      0
#> 572   0.685414533      0
#> 573  -0.570693210      0
#> 574  -0.446765724      0
#> 575   1.342688348      0
#> 576  -0.869634156      0
#> 577  -0.891031917      0
#> 578   0.605799797      0
#> 579   0.164068187      0
#> 580  -0.735584016      0
#> 581   0.646003324      0
#> 582  -0.649885108      0
#> 583   1.029918982      0
#> 584  -0.216800658      0
#> 585   1.312508830      0
#> 586   0.070946645      0
#> 587  -0.356743672      0
#> 588  -0.540014446      0
#> 589  -0.791240752      0
#> 590   0.180532254      0
#> 591  -0.340087276      0
#> 592   0.133819234      0
#> 593  -0.960918293      0
#> 594   0.671069803      0
#> 595   1.124845712      0
#> 596   0.683928581      0
#> 597  -0.160824333      0
#> 598   0.920482732      0
#> 599   0.045531058      0
#> 600   1.119790780      0
#> 601   1.240633824      0
#> 602   0.921671539      0
#> 603   0.476361513      0
#> 604   0.484116733      0
#> 605   0.075576902      0
#> 606   0.449345504      0
#> 607  -0.856885526      0
#> 608   0.353744058      0
#> 609  -1.252392735      0
#> 610   1.086868353      0
#> 611   1.792136044      0
#> 612   1.672460508      0
#> 613  -0.985454992      0
#> 614   0.214279252      0
#> 615   0.926218343      0
#> 616  -1.071817488      0
#> 617  -1.397707604      0
#> 618   0.408155562      0
#> 619  -0.363425217      0
#> 620   0.086224968      0
#> 621  -0.472576007      0
#> 622   0.975559080      0
#> 623   1.081413421      0
#> 624  -1.468645321      0
#> 625  -0.518790927      0
#> 626  -0.408992479      0
#> 627   0.891830940      0
#> 628   0.874342811      0
#> 629   0.816337338      0
#> 630  -1.381919488      0
#> 631   1.599680760      0
#> 632  -0.129034064      0
#> 633   0.962090442      0
#> 634   1.260095199      0
#> 635   0.251690221      0
#> 636  -1.943261979      0
#> 637   0.493668576      0
#> 638   0.366024223      0
#> 639  -0.449228533      0
#> 640  -0.859246745      0
#> 641  -0.940150578      0
#> 642  -0.016761991      0
#> 643  -0.482653793      0
#> 644  -1.333547580      0
#> 645   1.678016432      0
#> 646   0.085629169      0
#> 647  -0.707821621      0
#> 648  -0.842624988      0
#> 649  -0.209953532      0
#> 650  -1.291457203      0
#> 651  -0.593609653      0
#> 652   1.668873251      0
#> 653   0.345832212      0
#> 654   1.270609139      0
#> 655  -0.555957648      0
#> 656  -0.328579667      0
#> 657   1.196149835      0
#> 658   2.239455287      0
#> 659  -1.288653135      0
#> 660   0.591971803      0
#> 661   0.202138057      0
#> 662  -0.404120891      0
#> 663  -0.805190219      0
#> 664   1.025638041      0
#> 665   0.817569346      0
#> 666  -1.315302209      0
#> 667  -0.188528069      0
#> 668  -1.044889601      0
#> 669   0.624192210      0
#> 670  -0.078341389      0
#> 671  -0.634930488      0
#> 672  -0.472884837      0
#> 673  -1.099026642      0
#> 674  -0.259252057      0
#> 675  -0.484544006      0
#> 676  -0.399606453      0
#> 677  -1.231109973      0
#> 678   1.585079923      0
#> 679  -0.057726053      0
#> 680   0.074306198      0
#> 681  -0.053509725      0
#> 682   0.866753036      0
#> 683   0.749907589      0
#> 684  -1.028776530      0
#> 685  -0.533970539      0
#> 686  -2.089349092      0
#> 687   0.200019442      0
#> 688   0.703334415      0
#> 689  -0.329305595      0
#> 690   0.585166902      0
#> 691   1.333697886      0
#> 692  -0.185941697      0
#> 693  -0.249661972      0
#> 694   0.178101264      0
#> 695  -0.686278308      0
#> 696  -1.916517393      0
#> 697  -0.398793281      0
#> 698   0.575882343      0
#> 699  -1.260232914      0
#> 700   1.426912956      0
#> 701   1.763368352      0
#> 702   0.270334851      0
#> 703  -1.797002755      0
#> 704   0.653918140      0
#> 705  -0.776755203      0
#> 706  -0.978334865      0
#> 707   0.785721672      0
#> 708   0.658963005      0
#> 709   0.356279340      0
#> 710   1.238299891      0
#> 711   0.759180767      0
#> 712  -1.326787216      0
#> 713   1.024485535      0
#> 714   0.618030892      0
#> 715  -0.677124385      0
#> 716  -0.477324453      0
#> 717   0.803588181      0
#> 718  -0.021148500      0
#> 719  -1.115522776      0
#> 720  -0.656328456      0
#> 721   0.332586864      0
#> 722   1.367278808      0
#> 723   0.576188664      0
#> 724  -0.456647523      0
#> 725  -0.227040929      0
#> 726   0.549289984      0
#> 727   0.229825774      0
#> 728   0.849395268      0
#> 729  -0.413007386      0
#> 730  -1.427048670      0
#> 731   1.141756718      0
#> 732  -0.988993193      0
#> 733  -0.973203716      0
#> 734  -0.469586417      0
#> 735  -0.109947498      0
#> 736  -1.654355273      0
#> 737   0.362467915      0
#> 738  -0.349637399      0
#> 739  -0.143928517      0
#> 740  -0.650970382      0
#> 741   1.926739051      0
#> 742   0.066879066      0
#> 743  -0.980278987      0
#> 744   0.014485223      0
#> 745   0.592739678      0
#> 746  -0.923890576      0
#> 747  -0.205388555      0
#> 748   1.274475108      0
#> 749   0.535027516      0
#> 750   1.579540663      0
#> 751  -1.533508710      0
#> 752   0.036609779      0
#> 753   0.865391781      0
#> 754   0.492974452      0
#> 755   1.157973002      0
#> 756  -0.011628998      0
#> 757  -0.426013758      0
#> 758   0.706715793      0
#> 759  -0.075964778      0
#> 760   0.173424971      0
#> 761  -0.606670736      0
#> 762   0.792232727      0
#> 763  -0.982807685      0
#> 764  -0.443179204      0
#> 765   0.625381681      0
#> 766  -0.446501692      0
#> 767  -0.462194401      0
#> 768   0.032447641      0
#> 769  -0.253293977      0
#> 770  -2.107479115      0
#> 771  -0.128693849      0
#> 772  -0.649627042      0
#> 773   0.742338086      0
#> 774  -1.026654710      0
#> 775  -0.577820198      0
#> 776  -0.775534644      0
#> 777  -0.791805894      0
#> 778   0.607830222      0
#> 779   0.292823475      0
#> 780  -1.306481794      0
#> 781  -0.648422677      0
#> 782  -0.100698052      0
#> 783   0.554348857      0
#> 784   1.988155258      0
#> 785  -0.608678735      0
#> 786   0.156215048      0
#> 787   1.819129244      0
#> 788   1.137066493      0
#> 789  -1.407720481      0
#> 790  -0.912966536      0
#> 791   0.915267652      0
#> 792   0.099133844      0
#> 793  -1.823033423      0
#> 794  -1.274653211      0
#> 795  -0.521382605      0
#> 796  -0.391338336      0
#> 797  -0.938336046      0
#> 798  -0.391559548      0
#> 799   0.176974703      0
#> 800  -1.657096789      0
#> 801   0.329157743      0
#> 802  -0.818684984      0
#> 803  -0.831640571      0
#> 804   0.723449606      0
#> 805   0.946663278      0
#> 806   1.953692808      0
#> 807   1.234371163      0
#> 808   0.547567235      0
#> 809   0.071837033      0
#> 810  -1.503208011      0
#> 811   1.635438681      0
#> 812  -2.091174092      0
#> 813  -0.900530895      0
#> 814  -1.368489213      0
#> 815   0.285858849      0
#> 816   0.631979746      0
#> 817   0.531164700      0
#> 818   0.685318679      0
#> 819  -0.608507332      0
#> 820  -1.609431372      0
#> 821   0.414208021      0
#> 822   0.653327461      0
#> 823  -0.761834017      0
#> 824   1.387097351      0
#> 825  -0.943147200      0
#> 826  -0.608444125      0
#> 827  -0.021908879      0
#> 828   0.230501306      0
#> 829   2.112097467      0
#> 830  -0.082512879      0
#> 831  -1.634924085      0
#> 832  -0.155021819      0
#> 833   1.481087372      0
#> 834  -0.542113760      0
#> 835  -0.687442243      0
#> 836  -0.315032174      0
#> 837  -0.154670188      0
#> 838   0.227262271      0
#> 839   0.285493635      0
#> 840  -0.509925633      0
#> 841  -0.950762566      0
#> 842   0.930661706      0
#> 843   0.408700953      0
#> 844   0.434981426      0
#> 845   0.837080248      0
#> 846   0.679431327      0
#> 847   2.100240501      0
#> 848   1.895376684      0
#> 849   0.579219220      0
#> 850   0.188854951      0
#> 851   0.255141670      0
#> 852   0.030957234      0
#> 853   0.557856869      0
#> 854  -0.455210898      0
#> 855  -0.316552057      0
#> 856   1.793280515      0
#> 857  -1.298304397      0
#> 858   0.299611314      0
#> 859  -1.786081848      0
#> 860   0.565082812      0
#> 861   0.887700208      0
#> 862  -0.894072610      0
#> 863  -0.516606907      0
#> 864   1.501216682      0
#> 865  -0.681552340      0
#> 866   0.139387468      0
#> 867  -1.935606117      0
#> 868   0.780568352      0
#> 869   0.008576165      0
#> 870   0.503729490      0
#> 871  -0.872765311      0
#> 872   1.454576373      0
#> 873   0.515359154      0
#> 874   0.969056058      0
#> 875   0.191287244      0
#> 876  -1.308336445      0
#> 877   0.563604501      0
#> 878  -0.101105092      0
#> 879  -0.998361481      0
#> 880   1.191063032      0
#> 881   0.771308375      0
#> 882   0.347032513      0
#> 883   0.866899750      0
#> 884   1.348803800      0
#> 885   2.247730244      0
#> 886   0.605314089      0
#> 887  -0.919323555      0
#> 888   0.740514758      0
#> 889   1.684237916      0
#> 890  -0.349263003      0
#> 891  -2.482915016      0
#> 892   0.019645892      0
#> 893   0.790727346      0
#> 894   1.416299087      1
#> 895   0.150496756      1
#> 896   0.420316052      1
#> 897   0.742441002      1
#> 898   1.955367028      1
#> 899   0.581448749      1
#> 900  -0.273316070      1
#> 901   0.299600235      1
#> 902   0.492919571      1
#> 903   0.932276764      1
#> 904   0.785119737      1
#> 905   0.987195883      1
#> 906   1.231488468      1
#> 907   0.191394900      1
#> 908   1.291166924      1
#> 909   2.494146464      1
#> 910   0.551042695      1
#> 911   2.517483619      1
#> 912  -0.944228988      1
#> 913   0.086064880      1
#> 914  -1.439289968      1
#> 915   2.187369613      1
#> 916   1.338953865      1
#> 917   2.911599979      1
#> 918   0.577983330      1
#> 919   2.441562269      1
#> 920   1.580396233      1
#> 921   1.080238659      1
#> 922   1.166801273      1
#> 923   1.405113392      1
#> 924   0.273279765      1
#> 925   1.806387396      1
#> 926   0.644299207      1
#> 927   1.208427266      1
#> 928  -0.290965722      1
#> 929   1.834804126      1
#> 930   2.025708352      1
#> 931   0.551758006      1
#> 932  -0.526475752      1
#> 933   1.019225033      1
#> 934   0.291806990      1
#> 935  -0.328717113      1
#> 936   0.528548389      1
#> 937   0.979637422      1
#> 938   1.401542100      1
#> 939   0.505775546      1
#> 940   0.413290939      1
#> 941   1.003599725      1
#> 942   1.259152216      1
#> 943   1.199561096      1
#> 944   1.868947062      1
#> 945   2.299082932      1
#> 946   0.451674270      1
#> 947   0.036468768      1
#> 948   1.715461261      1
#> 949   0.697636031      1
#> 950  -1.095154006      1
#> 951  -1.103072649      1
#> 952  -2.483393422      1
#> 953   0.156776350      1
#> 954   1.980008537      1
#> 955  -0.156951113      1
#> 956   0.768263511      1
#> 957   0.221852806      1
#> 958   0.087237520      1
#> 959   1.819427733      1
#> 960  -0.535864952      1
#> 961   0.294267377      1
#> 962   1.527767687      1
#> 963   0.498672105      1
#> 964   1.996020197      1
#> 965   1.564616438      1
#> 966   0.181158415      1
#> 967   0.886175108      1
#> 968   1.277160899      1
#> 969   1.302297950      1
#> 970   0.469660326      1
#> 971   1.172207616      1
#> 972   0.697574204      1
#> 973   1.249557040      1
#> 974   0.355865658      1
#> 975  -0.362905580      1
#> 976   1.272082477      1
#> 977   0.252658422      1
#> 978   1.238721072      1
#> 979  -0.675961587      1
#> 980   1.344811947      1
#> 981   0.544321226      1
#> 982   0.002269246      1
#> 983   1.929099184      1
#> 984   0.712617566      1
#> 985   0.533838906      1
#> 986   0.788701144      1
#> 987  -0.138745307      1
#> 988   1.761043242      1
#> 989  -0.759583138      1
#> 990  -0.363211757      1
#> 991   1.048184280      1
#> 992   0.268395319      1
#> 993  -0.591509543      1
#> 994   0.901492747      1
#> 995   2.509977358      1
#> 996  -0.486203753      1
#> 997   0.847945602      1
#> 998   3.061477747      1
#> 999  -0.185980328      1
#> 1000 -0.194499333      1