Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -1.5362915037      0
#> 2    -1.3129167862      0
#> 3     0.3456281761      0
#> 4     1.1183026533      0
#> 5    -1.0680466753      0
#> 6    -1.7228519303      0
#> 7     0.5578087514      0
#> 8    -0.6618342468      0
#> 9     1.2818821722      0
#> 10   -1.0552934845      0
#> 11    0.6792947973      0
#> 12   -0.1343946309      0
#> 13    0.3937089667      0
#> 14   -0.9249843218      0
#> 15   -0.0628234459      0
#> 16    0.5890587770      0
#> 17    0.5564457402      0
#> 18   -0.0681394618      0
#> 19    2.5339332877      0
#> 20   -1.8412179956      0
#> 21    0.2648884608      0
#> 22   -1.8563346680      0
#> 23   -1.3314365282      0
#> 24    1.1447428673      0
#> 25   -0.4105015461      0
#> 26    0.7882417225      0
#> 27   -0.2438932979      0
#> 28   -1.5773942800      0
#> 29   -1.4458253410      0
#> 30    0.7937051310      0
#> 31   -0.6497227379      0
#> 32    2.9950908409      0
#> 33   -1.6086900513      0
#> 34   -0.0503471237      0
#> 35   -0.8997809246      0
#> 36    0.0083880548      0
#> 37   -1.5511929478      0
#> 38    0.3567437596      0
#> 39   -1.2404040014      0
#> 40   -0.8205775213      0
#> 41   -0.0229898947      0
#> 42   -0.4261149864      0
#> 43    0.7163249299      0
#> 44   -0.0752703182      0
#> 45   -1.3855555208      0
#> 46   -0.4949436565      0
#> 47    0.0287464785      0
#> 48    0.3204590242      0
#> 49    2.4718775119      0
#> 50   -1.1251032746      0
#> 51   -0.5082792749      0
#> 52   -0.1021400813      0
#> 53    0.7474006741      0
#> 54    1.0305198268      0
#> 55    1.2273744610      0
#> 56    0.8548887671      0
#> 57   -0.2813060008      0
#> 58   -0.6252054053      0
#> 59   -0.3732196754      0
#> 60    0.4243356555      0
#> 61   -0.2979557439      0
#> 62    0.7048572840      0
#> 63    0.6860796344      0
#> 64    0.4644913883      0
#> 65    0.1007940982      0
#> 66   -0.4420118928      0
#> 67    0.7908901146      0
#> 68   -0.2444184150      0
#> 69   -1.7267381631      0
#> 70   -1.4535420224      0
#> 71    0.4133215665      0
#> 72   -0.5745772302      0
#> 73    0.2145219467      0
#> 74    1.1556819963      0
#> 75   -1.9616197529      0
#> 76   -0.2161132743      0
#> 77   -1.3851766389      0
#> 78    0.1689354984      0
#> 79   -0.9798043099      0
#> 80   -0.8077142897      0
#> 81    0.3313260613      0
#> 82    0.2109503153      0
#> 83   -1.4127473801      0
#> 84   -1.1068484278      0
#> 85    1.2796188585      0
#> 86   -0.5307784836      0
#> 87   -0.1903209433      0
#> 88   -0.6329995521      0
#> 89    1.6054618461      0
#> 90   -0.5896279727      0
#> 91   -0.1022808475      0
#> 92   -0.5347258436      0
#> 93   -0.8834541994      0
#> 94    1.6301138061      0
#> 95    1.8834706561      0
#> 96    0.1911147061      0
#> 97    0.2516419986      0
#> 98    0.9787486234      0
#> 99   -1.6746674780      0
#> 100  -0.1062342314      0
#> 101   1.2796300922      0
#> 102   1.6973115403      0
#> 103   0.1120808804      0
#> 104  -0.3948611755      0
#> 105   0.5568853775      0
#> 106  -0.1512273524      0
#> 107   0.2485125615      0
#> 108   0.5334967621      0
#> 109  -0.2088580360      0
#> 110   2.4605068271      0
#> 111   0.1804573136      0
#> 112  -2.2849919946      0
#> 113   0.5517606389      0
#> 114  -1.5428019261      0
#> 115   1.6290893517      0
#> 116   1.1753181645      0
#> 117   0.0151244630      0
#> 118  -0.6823756926      0
#> 119   2.3988925886      0
#> 120   0.9527560998      0
#> 121   0.0583717524      0
#> 122   0.5694045238      0
#> 123  -0.4734231281      0
#> 124   0.4081309912      0
#> 125  -0.0486380737      0
#> 126  -0.4068512541      0
#> 127   0.1406255349      0
#> 128  -0.2967318257      0
#> 129  -0.7428532981      0
#> 130   0.3024088844      0
#> 131   0.2588592563      0
#> 132   0.3542352743      0
#> 133   1.1814075298      0
#> 134   1.3392574246      0
#> 135  -1.2399362983      0
#> 136   0.7425743537      0
#> 137  -2.4791488665      0
#> 138  -0.1829998324      0
#> 139  -1.4796890903      0
#> 140  -0.8596681736      0
#> 141   1.7738762659      0
#> 142  -0.2406394016      0
#> 143   0.5858643630      0
#> 144   0.5861976215      0
#> 145   0.4610748560      0
#> 146   0.2368483287      0
#> 147  -0.6423742738      0
#> 148  -1.5347059415      0
#> 149  -0.5772643311      0
#> 150  -0.3005963927      0
#> 151  -0.0597873202      0
#> 152  -0.4653497972      0
#> 153   1.4520029272      0
#> 154   0.2898189045      0
#> 155   0.5743298815      0
#> 156  -1.6465913337      0
#> 157   0.7972725993      0
#> 158  -0.3099336720      0
#> 159   0.4186675843      0
#> 160  -1.3595464834      0
#> 161   1.1146351521      0
#> 162  -1.6158462614      0
#> 163  -0.4480729246      0
#> 164   0.2401778404      0
#> 165   0.0800551557      0
#> 166  -0.6125199143      0
#> 167   0.0641193660      0
#> 168  -1.9259090745      0
#> 169   0.8379993197      0
#> 170  -0.6295231982      0
#> 171  -0.4941318886      0
#> 172   1.4967893746      0
#> 173  -0.2733967468      0
#> 174  -0.3689464289      0
#> 175  -0.9717962411      0
#> 176  -2.4691248574      0
#> 177  -0.7259322718      0
#> 178  -1.9280750166      0
#> 179   0.3685924868      0
#> 180   1.0012843870      0
#> 181  -0.9753840446      0
#> 182   0.0560464930      0
#> 183  -2.0348138951      0
#> 184   1.5225296465      0
#> 185   0.6573303943      0
#> 186   2.2985117947      0
#> 187   0.0001365607      0
#> 188   1.2710640560      0
#> 189  -0.9254762739      0
#> 190   1.8140359565      0
#> 191  -0.0088732990      0
#> 192  -0.7219269250      0
#> 193  -2.0690662030      0
#> 194  -1.6067540068      0
#> 195  -0.2264022352      0
#> 196   0.3005429022      0
#> 197   1.5520942196      0
#> 198  -0.4008724093      0
#> 199   0.8181334076      0
#> 200  -0.5448099774      0
#> 201  -1.2181481041      0
#> 202   0.0093252369      0
#> 203   2.3237480910      0
#> 204  -0.6260326182      0
#> 205   0.3150447209      0
#> 206   0.8665275487      0
#> 207  -0.1953914538      0
#> 208   0.6238422321      0
#> 209  -1.3997426687      0
#> 210  -0.9705825527      0
#> 211  -0.3828663788      0
#> 212  -0.1186378200      0
#> 213  -0.3170846005      0
#> 214   0.1013305780      0
#> 215  -0.9757870701      0
#> 216  -1.6841733549      0
#> 217  -0.2161690986      0
#> 218   0.4644834759      0
#> 219  -0.5700402542      0
#> 220  -0.3113751251      0
#> 221  -0.4980480843      0
#> 222   0.6957782817      0
#> 223   0.2607182521      0
#> 224   0.6270470427      0
#> 225  -1.7927282380      0
#> 226  -0.4670551652      0
#> 227   0.2455929520      0
#> 228   1.2457966647      0
#> 229   0.6318671416      0
#> 230   0.1014556171      0
#> 231   0.0628705013      0
#> 232  -0.5413880163      0
#> 233  -0.9239892807      0
#> 234   1.0321451217      0
#> 235   0.5847004921      0
#> 236  -0.8599592565      0
#> 237  -0.5289222642      0
#> 238   1.3386461086      0
#> 239  -1.0118983770      0
#> 240   0.9894945325      0
#> 241   0.3445710709      0
#> 242   2.2162928199      0
#> 243  -0.5059976411      0
#> 244  -0.0761628250      0
#> 245  -0.1291648036      0
#> 246   1.3173332211      0
#> 247   0.6347073964      0
#> 248  -0.4664690761      0
#> 249  -0.0126764184      0
#> 250   0.5789265440      0
#> 251  -1.5975298923      0
#> 252  -0.2592104294      0
#> 253  -0.8039299553      0
#> 254  -1.3193466115      0
#> 255   1.8255640750      0
#> 256  -1.6430374949      0
#> 257  -1.8104819080      0
#> 258  -1.8163415589      0
#> 259   0.3419992021      0
#> 260   1.0440658165      0
#> 261   0.0339519786      0
#> 262   0.7210985880      0
#> 263   0.7182526807      0
#> 264   1.3839411625      0
#> 265   0.2446419601      0
#> 266  -1.6398593712      0
#> 267   0.3518559862      0
#> 268  -0.9893810812      0
#> 269   0.7629227067      0
#> 270  -2.1338111554      0
#> 271   0.1089316909      0
#> 272  -0.1642431683      0
#> 273   0.5843578042      0
#> 274   0.9100466525      0
#> 275   2.0275733987      0
#> 276  -0.1478259865      0
#> 277  -0.6558543638      0
#> 278  -0.3772312816      0
#> 279   0.0398598681      0
#> 280   0.9337670249      0
#> 281   0.5737516992      0
#> 282   1.2024572322      0
#> 283  -0.8636943830      0
#> 284  -0.1751763042      0
#> 285  -0.3971829423      0
#> 286  -2.0164913137      0
#> 287  -0.1259759359      0
#> 288   0.9396726999      0
#> 289  -0.9171455094      0
#> 290  -0.4875862851      0
#> 291  -1.0898829031      0
#> 292  -1.4033328726      0
#> 293  -1.6732715534      0
#> 294  -0.3637590556      0
#> 295  -0.0142860158      0
#> 296   0.1863923822      0
#> 297  -0.9908224662      0
#> 298   1.2784499211      0
#> 299   0.2954558086      0
#> 300  -0.2715842319      0
#> 301   0.3313589952      0
#> 302   0.1858761537      0
#> 303  -0.7321480539      0
#> 304   1.3280946240      0
#> 305  -1.4796343744      0
#> 306   0.5159449391      0
#> 307  -1.1235957300      0
#> 308  -0.8217861028      0
#> 309   0.4742646934      0
#> 310  -0.3377284651      0
#> 311  -0.5578921900      0
#> 312  -0.3842654804      0
#> 313   0.3888956668      0
#> 314  -0.6082582146      0
#> 315   0.2457785139      0
#> 316   0.9505244459      0
#> 317   0.5498297470      0
#> 318  -0.9334551805      0
#> 319  -0.8416133517      0
#> 320   1.0515398142      0
#> 321   0.7430007646      0
#> 322  -1.4927904625      0
#> 323   0.9051785459      0
#> 324  -0.1058103413      0
#> 325  -0.4384385814      0
#> 326  -1.7163854556      0
#> 327   0.3364387209      0
#> 328   0.5001427145      0
#> 329   0.9431336555      0
#> 330  -0.1112168495      0
#> 331   3.0025067015      0
#> 332  -0.2365542280      0
#> 333   0.3625374844      0
#> 334  -0.2337633761      0
#> 335  -0.0212325248      0
#> 336   0.9626650365      0
#> 337  -0.4131284011      0
#> 338   0.1243968829      0
#> 339   1.8310206547      0
#> 340  -1.4533372850      0
#> 341  -2.6559863398      0
#> 342   1.1177308401      0
#> 343   0.6777280276      0
#> 344   2.4975873971      0
#> 345   1.4413136463      0
#> 346   0.1921273114      0
#> 347   0.1274275884      0
#> 348  -1.7760882042      0
#> 349   1.8089771388      0
#> 350  -1.8278858788      0
#> 351  -0.3670690360      0
#> 352  -0.9836053901      0
#> 353   2.1335496627      0
#> 354   1.3263521899      0
#> 355  -0.2945573911      0
#> 356   0.3277200352      0
#> 357  -0.7892659816      0
#> 358  -0.1490361217      0
#> 359  -0.5481327953      0
#> 360   1.1336314094      0
#> 361  -0.5398058286      0
#> 362   0.6321314838      0
#> 363   0.6895288289      0
#> 364   1.1643022527      0
#> 365  -0.4212142220      0
#> 366   0.2821103107      0
#> 367  -0.6332106496      0
#> 368  -0.0285769065      0
#> 369   0.8280754554      0
#> 370  -1.9837928871      0
#> 371   0.4320455524      0
#> 372   0.0447454928      0
#> 373  -0.5617823765      0
#> 374  -0.7806568909      0
#> 375  -0.2044971420      0
#> 376   0.1848675037      0
#> 377  -1.7640712362      0
#> 378   0.9045493275      0
#> 379   0.1136412642      0
#> 380   1.7818911037      0
#> 381   0.1140639899      0
#> 382   0.2585415413      0
#> 383  -0.0284115447      0
#> 384   0.5525892408      0
#> 385   0.7741836481      0
#> 386  -1.5003366661      0
#> 387  -0.7038173838      0
#> 388  -0.0616023809      0
#> 389   1.5215677435      0
#> 390   0.5953541780      0
#> 391  -0.6114101072      0
#> 392   0.2202570713      0
#> 393  -0.2657075939      0
#> 394  -0.8135330006      0
#> 395   0.3650572589      0
#> 396  -0.1351549098      0
#> 397   1.2882064980      0
#> 398  -2.1092962087      0
#> 399   1.1137122593      0
#> 400   0.6637082182      0
#> 401   1.0240320855      0
#> 402  -1.5656905638      0
#> 403   0.3254404461      0
#> 404  -0.9180205850      0
#> 405  -0.7698006975      0
#> 406   1.3045452404      0
#> 407  -0.9154600176      0
#> 408  -0.6114983093      0
#> 409   1.6259072436      0
#> 410  -0.0205544075      0
#> 411   0.3227255159      0
#> 412  -0.0650702686      0
#> 413  -1.1815593325      0
#> 414   1.1994481286      0
#> 415   0.4462911206      0
#> 416  -0.8395020925      0
#> 417   0.3775472305      0
#> 418   0.9215405750      0
#> 419   1.1889196215      0
#> 420   0.1028108292      0
#> 421   0.9528647712      0
#> 422   0.8895711544      0
#> 423   0.4179781214      0
#> 424  -0.9489654949      0
#> 425  -1.4896030455      0
#> 426  -0.1973426437      0
#> 427  -1.9857361386      0
#> 428   0.7638252096      0
#> 429  -0.4502004203      0
#> 430   1.2591229501      0
#> 431   0.4347188228      0
#> 432   0.6512193616      0
#> 433  -1.6669144589      0
#> 434   0.7978481120      0
#> 435   0.6658881252      0
#> 436  -1.0549344715      0
#> 437   1.1059022211      0
#> 438   0.7710826903      0
#> 439  -0.9561771001      0
#> 440   0.3807879001      0
#> 441   0.7459893101      0
#> 442  -0.0674780677      0
#> 443  -0.4551440152      0
#> 444   0.8392453479      0
#> 445  -2.5221543379      0
#> 446   0.3009750298      0
#> 447   0.1555406403      0
#> 448  -1.7574397699      0
#> 449  -0.1046006493      0
#> 450  -1.1437955164      0
#> 451   0.4152274164      0
#> 452  -0.5432467914      0
#> 453  -0.7808132761      0
#> 454   2.0276104439      0
#> 455   0.3729725599      0
#> 456  -0.5734970055      0
#> 457   0.6821418371      0
#> 458   0.8846739295      0
#> 459  -0.4197194098      0
#> 460   0.5163424978      0
#> 461  -1.3568867840      0
#> 462   1.0965484597      0
#> 463   0.3760085542      0
#> 464   1.0295234555      0
#> 465  -0.1540980586      0
#> 466  -0.3206122788      0
#> 467  -0.6213646154      0
#> 468  -0.8181733437      0
#> 469   0.1368201021      0
#> 470   0.5286907122      0
#> 471  -0.9938982810      0
#> 472   1.2136433367      0
#> 473  -0.2959956959      0
#> 474   0.7141846867      0
#> 475   0.1464360415      0
#> 476  -0.4524997955      0
#> 477  -0.3348289863      0
#> 478   0.2113448020      0
#> 479  -1.4199440231      0
#> 480  -0.8762269611      0
#> 481  -1.7697814944      0
#> 482  -0.8321212209      0
#> 483   0.0947001394      0
#> 484   1.6470826528      0
#> 485  -0.0589320693      0
#> 486   0.1117515321      0
#> 487   0.1313451687      0
#> 488   0.2897261782      0
#> 489  -0.0670480123      0
#> 490   0.3749911504      0
#> 491  -1.4433042433      0
#> 492   0.0531493171      0
#> 493   0.6413910765      0
#> 494   0.8010340276      0
#> 495  -0.7188368929      0
#> 496   0.0339499073      0
#> 497  -0.4821461577      0
#> 498  -0.2873916847      0
#> 499   0.3518166956      0
#> 500   0.7660873687      0
#> 501  -0.1897770788      0
#> 502  -0.7542224092      0
#> 503  -0.2184569379      0
#> 504  -0.2357478027      0
#> 505   0.2364163552      0
#> 506   0.6556943899      0
#> 507  -1.6408562990      0
#> 508  -0.0866164641      0
#> 509   1.2503161420      0
#> 510   0.7523328595      0
#> 511   1.5176730026      0
#> 512  -1.3731656770      0
#> 513  -1.0843693385      0
#> 514   1.4246774241      0
#> 515   0.2879359960      0
#> 516   2.2268972629      0
#> 517   1.4325392471      0
#> 518  -0.5142528806      0
#> 519   0.2064381787      0
#> 520  -1.8847392398      0
#> 521  -1.2566817564      0
#> 522  -1.3514647608      0
#> 523  -1.1162312632      0
#> 524   0.0913334264      0
#> 525   0.0375776107      0
#> 526  -0.6110655822      0
#> 527  -1.0325640284      0
#> 528   0.4649739572      0
#> 529   0.5337569255      0
#> 530   0.8236249609      0
#> 531  -0.9266407039      0
#> 532  -1.4758109647      0
#> 533   0.5418002892      0
#> 534   0.3996164084      0
#> 535   0.4497622511      0
#> 536   1.3441625989      0
#> 537  -1.4543985290      0
#> 538  -0.0721239182      0
#> 539   1.7713151020      0
#> 540   1.2314992923      0
#> 541   0.6662416120      0
#> 542  -0.3149794943      0
#> 543  -2.8915055669      0
#> 544  -0.5415288290      0
#> 545   0.5833646610      0
#> 546  -0.2829277352      0
#> 547   0.3146694915      0
#> 548  -0.9793702699      0
#> 549  -1.3039292040      0
#> 550   0.1733329625      0
#> 551   0.7032641295      0
#> 552   0.4888959425      0
#> 553   0.0748129931      0
#> 554   0.2426786147      0
#> 555   0.0680917669      0
#> 556   0.0632912407      0
#> 557   0.0821314170      0
#> 558  -0.3703475212      0
#> 559  -0.4026558962      0
#> 560   0.1827165236      0
#> 561  -0.0191221882      0
#> 562  -1.0083516823      0
#> 563   1.7969966624      0
#> 564  -0.8206784839      0
#> 565  -0.5886068016      0
#> 566   0.7562926727      0
#> 567   1.1954792510      0
#> 568  -0.3942017179      0
#> 569   0.8595054738      0
#> 570  -0.4441395972      0
#> 571   0.5791379508      0
#> 572  -0.2391227278      0
#> 573   1.8495701971      0
#> 574   0.4852160471      0
#> 575   1.3324028187      0
#> 576   1.7197484851      0
#> 577   0.1658496174      0
#> 578   0.1811822759      0
#> 579  -1.7646236860      0
#> 580  -0.7593538187      0
#> 581  -0.7624227057      0
#> 582  -0.9671762978      0
#> 583   0.4733294009      0
#> 584  -0.1723499672      0
#> 585  -0.3375320057      0
#> 586  -0.0555121554      0
#> 587  -0.3343591235      0
#> 588  -1.9395903125      0
#> 589  -0.3073433837      0
#> 590  -0.1669360255      0
#> 591  -1.4724715178      0
#> 592   0.0720952435      0
#> 593   0.2293622692      0
#> 594  -0.6754557244      0
#> 595  -0.3526221073      0
#> 596  -0.4341156933      0
#> 597   0.6607938248      0
#> 598  -0.3081374125      0
#> 599  -1.1344930017      0
#> 600  -0.7160288646      0
#> 601  -1.7301132799      0
#> 602  -0.6130398084      0
#> 603  -0.4055119088      0
#> 604   0.0357895940      0
#> 605   0.9986682394      0
#> 606   0.2398514668      0
#> 607   1.6364410054      0
#> 608   0.0283973688      0
#> 609   0.3833090541      0
#> 610   0.1954815601      0
#> 611  -1.6276797741      0
#> 612   0.6662329248      0
#> 613   1.0142967700      0
#> 614   0.5024626068      0
#> 615   0.5481302040      0
#> 616  -0.5905651751      0
#> 617  -0.0925464558      0
#> 618   0.7739623900      0
#> 619   0.1757141073      0
#> 620  -1.4845677688      0
#> 621   0.1799213972      0
#> 622   0.8546671622      0
#> 623  -0.7518163366      0
#> 624  -1.4748831588      0
#> 625   0.5641470469      0
#> 626   2.9747660472      0
#> 627  -0.8964871052      0
#> 628  -0.2627010960      0
#> 629  -0.0024341873      0
#> 630   0.8905361122      0
#> 631   0.3007755819      0
#> 632   0.0943179113      0
#> 633   0.0301396197      0
#> 634  -1.9418817327      0
#> 635  -0.8144973751      0
#> 636   0.1923938274      0
#> 637  -1.2467738987      0
#> 638   0.0288948459      0
#> 639  -0.0756294347      0
#> 640  -0.9938317194      0
#> 641  -0.8293098193      0
#> 642   0.5657764702      0
#> 643  -0.6565271625      0
#> 644   0.9631031190      0
#> 645  -0.1128420258      0
#> 646  -0.9678810161      0
#> 647  -2.5062124873      0
#> 648   0.4624174212      0
#> 649   1.5089116152      0
#> 650  -1.2005527448      0
#> 651   0.1822173012      0
#> 652   0.4145694073      0
#> 653   0.5874624317      0
#> 654   1.2387947700      0
#> 655  -0.3416623676      0
#> 656  -1.5298294814      0
#> 657   0.1528845300      0
#> 658   0.8211242272      0
#> 659  -0.0703133111      0
#> 660   0.0877844666      0
#> 661   0.1208435502      0
#> 662  -0.4879281429      0
#> 663   1.0481285454      0
#> 664  -1.2444366292      0
#> 665   1.5777854703      0
#> 666  -2.2703828719      0
#> 667   0.6111026183      0
#> 668   1.5715726838      0
#> 669   0.1112624152      0
#> 670  -0.0157104921      0
#> 671  -1.3933923610      0
#> 672   2.0314833658      0
#> 673  -0.9801149454      0
#> 674  -0.4894195283      0
#> 675  -0.3966529013      0
#> 676   1.2342731858      0
#> 677  -0.3361639427      0
#> 678  -1.7669053798      0
#> 679   0.1917137705      0
#> 680   0.1133359924      0
#> 681   0.7260710820      0
#> 682  -0.5094490448      0
#> 683  -1.1274992022      0
#> 684   1.3911190937      0
#> 685  -0.4100264430      0
#> 686  -1.0245551140      0
#> 687   1.0111495973      0
#> 688  -0.8473789598      0
#> 689  -0.1922870430      0
#> 690  -0.3800424657      0
#> 691   0.2171283991      0
#> 692  -0.1920105786      0
#> 693  -0.1177945835      0
#> 694  -2.2896836054      0
#> 695  -0.3918103697      0
#> 696   1.1668464995      0
#> 697  -0.0505351945      0
#> 698   1.3434140738      0
#> 699   0.0159326803      0
#> 700   0.4969971824      0
#> 701  -0.4887698017      0
#> 702  -2.0490203000      0
#> 703  -0.6859654622      0
#> 704  -0.7076596257      0
#> 705  -0.3333176458      0
#> 706   0.1126037941      0
#> 707   0.1768722246      0
#> 708   0.1978813793      0
#> 709   0.2535404077      0
#> 710  -0.6209705765      0
#> 711   0.1709730946      0
#> 712  -0.4645649491      0
#> 713  -0.3057996186      0
#> 714   0.6965532939      0
#> 715  -1.2771923399      0
#> 716  -1.6161587961      0
#> 717   0.8800480897      0
#> 718   0.9723324690      0
#> 719   0.8808931366      0
#> 720  -0.3596419131      0
#> 721  -0.0965657103      0
#> 722  -0.2916834901      0
#> 723  -0.5251380229      0
#> 724  -0.0601592670      0
#> 725  -1.7765509008      0
#> 726  -1.7010583164      0
#> 727   0.8084959335      0
#> 728  -0.5836664478      0
#> 729   0.3365153589      0
#> 730   0.8506831613      0
#> 731   0.8931847124      0
#> 732   0.0781979555      0
#> 733  -0.6267438723      0
#> 734  -0.9177589148      0
#> 735  -0.5915046726      0
#> 736  -0.4747827438      0
#> 737   0.8228611413      0
#> 738  -0.0352557585      0
#> 739   2.3303935522      0
#> 740   0.4565473548      0
#> 741   1.0811291787      0
#> 742  -0.2214871666      0
#> 743  -2.2074751643      0
#> 744   0.2015095254      0
#> 745  -0.4641959357      0
#> 746   0.5493949164      0
#> 747   1.3723580795      0
#> 748  -1.3952368032      0
#> 749  -0.0227747261      0
#> 750   1.5046578449      0
#> 751  -0.1462761067      0
#> 752   0.2771071588      0
#> 753  -0.0930605535      0
#> 754   1.1213945738      0
#> 755  -1.6740782618      0
#> 756   0.4574575271      0
#> 757  -0.2420978796      0
#> 758   0.4708879911      0
#> 759   1.8761486564      0
#> 760  -0.3886491130      0
#> 761  -0.8464188424      0
#> 762  -0.2718424224      0
#> 763   1.4706230096      0
#> 764  -0.4416444249      0
#> 765   0.1249549000      0
#> 766  -0.3708375299      0
#> 767   1.1496403640      0
#> 768   0.2565556111      0
#> 769   1.2608724881      0
#> 770   0.7292882137      0
#> 771  -1.9373320817      0
#> 772   0.2709959344      0
#> 773   0.7311347859      0
#> 774   0.2512894849      0
#> 775   0.3209701158      0
#> 776  -0.9732081744      0
#> 777   1.3213788435      0
#> 778  -0.1445454687      0
#> 779  -0.2673607736      0
#> 780  -2.2321494426      0
#> 781   0.4989818544      0
#> 782   0.4152173241      0
#> 783   0.8552976315      0
#> 784  -0.7177182995      0
#> 785   0.9481644799      0
#> 786  -1.0245811838      0
#> 787   1.3904861827      0
#> 788   0.0043816013      0
#> 789  -0.7876930668      0
#> 790  -0.3806219856      0
#> 791  -0.9382760931      0
#> 792  -1.1921304244      0
#> 793   0.2998780786      0
#> 794  -2.6260911350      0
#> 795   0.6106697541      0
#> 796   0.0849232725      0
#> 797  -0.6522216275      0
#> 798   0.4921084948      0
#> 799   0.1369691898      0
#> 800  -0.0680656064      0
#> 801  -0.6938375332      0
#> 802  -1.3059671558      0
#> 803  -1.4731743408      0
#> 804   1.0368154694      0
#> 805  -1.8188328097      0
#> 806  -0.6655703057      0
#> 807   0.1310027866      0
#> 808  -0.1608900912      0
#> 809   0.1418597666      0
#> 810  -2.5181728699      0
#> 811  -0.0046116216      0
#> 812  -0.5321658837      0
#> 813   0.5911663267      0
#> 814   1.6276300573      0
#> 815  -0.5250128410      0
#> 816  -1.0580834184      0
#> 817  -0.7131042062      0
#> 818   0.1314677505      0
#> 819   0.9443939521      0
#> 820   2.7339149361      0
#> 821   0.3302221198      0
#> 822  -0.5269497075      0
#> 823  -0.3333457073      0
#> 824   1.7098970887      0
#> 825   0.9134848664      0
#> 826   2.0619945668      0
#> 827   0.0885970053      0
#> 828  -0.5001837061      0
#> 829   0.5486564658      0
#> 830   0.5916804405      0
#> 831  -0.8056621463      0
#> 832  -0.7526028647      0
#> 833  -0.5682199032      0
#> 834  -0.8732424897      0
#> 835   0.3245446706      0
#> 836   0.9781437403      0
#> 837   0.2108125474      0
#> 838  -1.7397894094      0
#> 839  -0.8756787114      0
#> 840   0.6890053403      0
#> 841  -2.1925712016      0
#> 842   2.5433061376      0
#> 843  -0.9464593442      0
#> 844  -0.6052105045      0
#> 845   2.0098054914      0
#> 846   1.0911218510      0
#> 847  -0.4218164496      0
#> 848  -0.9343535294      0
#> 849  -0.3490279710      0
#> 850   0.0129318057      0
#> 851   1.0276688115      0
#> 852  -0.1436026677      0
#> 853   0.0384291516      0
#> 854   1.5425121217      0
#> 855  -0.7937134890      0
#> 856  -1.0706442825      0
#> 857  -0.3076549622      0
#> 858  -1.0321935315      0
#> 859   0.2637356930      0
#> 860  -0.9290443245      0
#> 861  -0.9596322373      0
#> 862  -1.0850795892      0
#> 863  -1.4763314495      0
#> 864   0.9632146372      0
#> 865   1.3131991018      0
#> 866  -1.8784497748      0
#> 867  -0.2426619414      0
#> 868  -0.2852206314      0
#> 869  -0.3918575552      0
#> 870  -0.0541188230      0
#> 871  -0.5874745810      0
#> 872   0.7380296227      0
#> 873   0.6153407966      0
#> 874  -0.1813399977      0
#> 875  -1.1481473188      0
#> 876   1.4430875961      0
#> 877  -0.7436710657      0
#> 878  -0.5288601542      0
#> 879  -1.1734540671      0
#> 880  -0.4434985182      0
#> 881  -0.4461563670      0
#> 882   0.1003933791      0
#> 883  -0.0687572876      0
#> 884   1.1182019123      0
#> 885  -0.0990492907      0
#> 886   0.1739527778      0
#> 887   0.7189641326      0
#> 888   1.1096411840      0
#> 889   0.4808713110      0
#> 890  -0.5478410561      0
#> 891   0.6200466392      0
#> 892  -1.0309903895      0
#> 893   1.6510984896      0
#> 894  -1.3904940045      0
#> 895  -0.3817766016      0
#> 896   0.6753964560      0
#> 897   2.2870611115      0
#> 898  -0.5287217935      0
#> 899   2.4517573099      0
#> 900   0.7590535842      0
#> 901   1.0344271508      0
#> 902   0.8745899703      0
#> 903  -0.3649405116      0
#> 904   1.5707597510      1
#> 905   1.0490708125      1
#> 906   0.5375115038      1
#> 907   0.6917038176      1
#> 908   0.9453778362      1
#> 909   0.0356043796      1
#> 910   0.5867166675      1
#> 911   1.0261943604      1
#> 912   0.1749022431      1
#> 913   0.5604349834      1
#> 914   0.8991193159      1
#> 915   0.6062626640      1
#> 916   1.9353908713      1
#> 917   0.5450780319      1
#> 918   1.0319749413      1
#> 919   0.7057101243      1
#> 920  -1.5346831307      1
#> 921   0.9867893711      1
#> 922  -0.0226860427      1
#> 923   0.3778405538      1
#> 924   1.5460152754      1
#> 925   2.3162917268      1
#> 926   0.1382337424      1
#> 927  -0.7173810890      1
#> 928   0.4811318199      1
#> 929  -0.2147530943      1
#> 930  -0.5955727009      1
#> 931  -0.9538360717      1
#> 932  -0.3326982513      1
#> 933   0.6866774466      1
#> 934   0.9029215039      1
#> 935  -0.4722657496      1
#> 936   1.4084923755      1
#> 937   0.4417769360      1
#> 938   0.3268611706      1
#> 939   1.4128086805      1
#> 940  -0.6821160207      1
#> 941   0.1869852087      1
#> 942   0.4227041780      1
#> 943   1.2821485144      1
#> 944  -0.1473008532      1
#> 945  -0.6146759184      1
#> 946  -0.0888734786      1
#> 947   0.7259207291      1
#> 948   0.6122189616      1
#> 949  -0.8104091510      1
#> 950   0.6440645004      1
#> 951   1.3556823646      1
#> 952  -0.0022019309      1
#> 953  -0.2134868008      1
#> 954  -0.3559726782      1
#> 955   1.4774385719      1
#> 956   0.8255701844      1
#> 957  -0.9688708129      1
#> 958  -0.8235911998      1
#> 959   1.3114318145      1
#> 960   0.2545400493      1
#> 961  -0.7864744641      1
#> 962   1.1540093816      1
#> 963   1.7119769691      1
#> 964   0.8534323676      1
#> 965   2.3240177935      1
#> 966   1.1409407653      1
#> 967   1.5384701524      1
#> 968   0.3012422906      1
#> 969   0.6976386905      1
#> 970   1.0604391285      1
#> 971   0.0947825699      1
#> 972  -0.6959572931      1
#> 973   0.9953946340      1
#> 974   1.4328548627      1
#> 975   0.1270318876      1
#> 976   1.5777442959      1
#> 977   1.4983960761      1
#> 978  -1.2921603533      1
#> 979   0.4693193472      1
#> 980   2.0417418945      1
#> 981   1.4209563699      1
#> 982  -0.2780493369      1
#> 983   0.8130230992      1
#> 984   0.8447265428      1
#> 985   1.2025837402      1
#> 986   0.2515950915      1
#> 987   0.0956120673      1
#> 988   2.4064068967      1
#> 989   0.1173973842      1
#> 990   2.4109230106      1
#> 991   0.4735923941      1
#> 992   1.8267522137      1
#> 993   0.3557887767      1
#> 994   1.0919352729      1
#> 995   0.3494169838      1
#> 996   0.7700380927      1
#> 997   0.1775604123      1
#> 998   0.7051752154      1
#> 999   1.4312177397      1
#> 1000  0.6965211808      1