Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.871521682      0
#> 2    -0.078238284      0
#> 3    -0.881268794      0
#> 4    -1.578224575      0
#> 5    -0.991511303      0
#> 6     0.039022237      0
#> 7    -0.204952171      0
#> 8    -0.398897512      0
#> 9     0.305679435      0
#> 10   -0.438968856      0
#> 11   -1.225109336      0
#> 12   -1.121962895      0
#> 13    1.772078989      0
#> 14    1.092585331      0
#> 15    0.220577920      0
#> 16   -0.470753562      0
#> 17    0.180940689      0
#> 18    0.368858126      0
#> 19   -0.238042077      0
#> 20   -0.621729835      0
#> 21    1.791538353      0
#> 22    0.635282901      0
#> 23   -0.849056356      0
#> 24    0.136451643      0
#> 25    0.705334204      0
#> 26   -1.152965804      0
#> 27    0.104381306      0
#> 28   -0.274691794      0
#> 29   -0.358533302      0
#> 30    2.246097036      0
#> 31   -0.232662745      0
#> 32    2.136031268      0
#> 33   -0.222616015      0
#> 34   -1.381177115      0
#> 35   -1.004087945      0
#> 36   -0.217807678      0
#> 37   -0.663288116      0
#> 38   -1.298809124      0
#> 39   -0.342910884      0
#> 40    0.416289738      0
#> 41   -2.675045143      0
#> 42   -0.625294895      0
#> 43    2.126803155      0
#> 44   -1.772029973      0
#> 45   -0.264743044      0
#> 46    0.713237286      0
#> 47   -0.270032590      0
#> 48   -1.337344953      0
#> 49    0.763691175      0
#> 50    0.024916745      0
#> 51    1.016463042      0
#> 52    1.286418460      0
#> 53   -0.400865692      0
#> 54    0.265654147      0
#> 55    1.426225550      0
#> 56    0.358143616      0
#> 57   -0.147980695      0
#> 58   -0.408059766      0
#> 59   -0.114027812      0
#> 60    0.531374043      0
#> 61   -1.223693539      0
#> 62    0.752168580      0
#> 63   -0.241659152      0
#> 64   -1.038471100      0
#> 65    0.317074026      0
#> 66    0.705684780      0
#> 67    0.301197561      0
#> 68    0.726846533      0
#> 69   -0.886733443      0
#> 70   -0.319609832      0
#> 71    0.097341076      0
#> 72   -1.459967718      0
#> 73    1.369077284      0
#> 74    0.156798800      0
#> 75    1.345429310      0
#> 76   -0.209546872      0
#> 77   -1.689887081      0
#> 78   -0.248191082      0
#> 79   -0.265658105      0
#> 80    1.758346299      0
#> 81   -1.153721176      0
#> 82    0.530981859      0
#> 83   -0.219961186      0
#> 84    0.777349853      0
#> 85   -0.913494095      0
#> 86    0.041905807      0
#> 87   -1.455503834      0
#> 88    1.601236878      0
#> 89    0.583448142      0
#> 90   -1.644477887      0
#> 91    0.166551720      0
#> 92   -0.504679854      0
#> 93   -0.311487997      0
#> 94   -2.198124375      0
#> 95   -0.195126707      0
#> 96   -1.371263237      0
#> 97   -3.272790932      0
#> 98    0.273100685      0
#> 99   -0.409590312      0
#> 100   0.818740342      0
#> 101  -1.505120053      0
#> 102   1.034358360      0
#> 103  -0.290455117      0
#> 104   0.852067746      0
#> 105  -2.217169903      0
#> 106  -2.112954487      0
#> 107  -0.816521661      0
#> 108  -0.108051552      0
#> 109   0.465189751      0
#> 110  -0.799182792      0
#> 111   0.239637592      0
#> 112  -0.012340641      0
#> 113   1.193990708      0
#> 114  -1.583668837      0
#> 115   2.567962866      0
#> 116   0.745543745      0
#> 117   0.351706607      0
#> 118   0.178828730      0
#> 119   0.242520575      0
#> 120   0.259051590      0
#> 121   0.690630327      0
#> 122   0.757585853      0
#> 123  -0.002951471      0
#> 124  -1.686370870      0
#> 125  -1.988440866      0
#> 126   0.539512277      0
#> 127   1.039402403      0
#> 128   2.172987678      0
#> 129   0.174901721      0
#> 130  -0.630822300      0
#> 131  -0.542071171      0
#> 132  -1.212339709      0
#> 133  -0.191111516      0
#> 134   1.689975775      0
#> 135  -1.023422080      0
#> 136  -0.003304015      0
#> 137  -2.132635893      0
#> 138   1.209121179      0
#> 139  -0.654379765      0
#> 140  -0.171374084      0
#> 141  -1.585740947      0
#> 142  -0.920345556      0
#> 143  -0.493932451      0
#> 144  -0.095666790      0
#> 145  -0.579482887      0
#> 146  -2.538169431      0
#> 147   0.365534826      0
#> 148  -0.104900276      0
#> 149   0.691740316      0
#> 150  -1.227749902      0
#> 151  -1.909765491      0
#> 152  -0.983916939      0
#> 153   0.912836239      0
#> 154   1.122902077      0
#> 155  -1.121166169      0
#> 156  -0.032373913      0
#> 157   2.012099574      0
#> 158   0.650034851      0
#> 159   0.816175576      0
#> 160   0.238396045      0
#> 161  -0.475927795      0
#> 162  -0.386309406      0
#> 163   0.696296538      0
#> 164  -0.844956507      0
#> 165  -0.344186415      0
#> 166  -0.376278553      0
#> 167  -0.995206586      0
#> 168   2.105699406      0
#> 169   1.529098776      0
#> 170  -0.759265606      0
#> 171   1.144192055      0
#> 172  -1.846408049      0
#> 173  -0.498975175      0
#> 174   1.397711433      0
#> 175   1.525037727      0
#> 176  -0.838974674      0
#> 177  -0.030670009      0
#> 178   0.824576278      0
#> 179  -0.312308759      0
#> 180  -0.401775931      0
#> 181  -0.133428264      0
#> 182  -0.300421019      0
#> 183  -0.612113751      0
#> 184   2.376045600      0
#> 185  -0.195581853      0
#> 186  -1.341275632      0
#> 187  -0.817419345      0
#> 188  -0.617433335      0
#> 189  -0.700626432      0
#> 190   0.704212214      0
#> 191  -1.481516999      0
#> 192  -0.730582166      0
#> 193   0.067810814      0
#> 194   0.612590410      0
#> 195   0.931151253      0
#> 196   0.389146262      0
#> 197   0.280280158      0
#> 198   0.770022251      0
#> 199   0.600789844      0
#> 200  -1.015798806      0
#> 201   0.443148593      0
#> 202   1.247706525      0
#> 203  -0.685820530      0
#> 204  -2.037088307      0
#> 205  -0.209686877      0
#> 206   2.314785413      0
#> 207  -0.426008583      0
#> 208   0.146386207      0
#> 209   1.656374921      0
#> 210  -0.825553904      0
#> 211  -0.629840194      0
#> 212   0.791248820      0
#> 213   0.111008513      0
#> 214   0.428135194      0
#> 215  -1.160968564      0
#> 216  -1.465861765      0
#> 217   0.400605736      0
#> 218   1.169101678      0
#> 219  -0.159598035      0
#> 220  -1.939032854      0
#> 221   0.914638587      0
#> 222  -1.254061859      0
#> 223   0.029406354      0
#> 224   0.878503840      0
#> 225  -2.328036932      0
#> 226   0.427241726      0
#> 227   1.109567598      0
#> 228   1.341868759      0
#> 229  -0.707274320      0
#> 230   1.150049673      0
#> 231  -0.619196641      0
#> 232  -0.134963550      0
#> 233   1.159598170      0
#> 234   0.344617584      0
#> 235   0.682186599      0
#> 236   0.741990664      0
#> 237  -0.512160680      0
#> 238   0.268067476      0
#> 239   0.676216799      0
#> 240   0.191911191      0
#> 241  -1.494741421      0
#> 242  -0.165352804      0
#> 243   0.105472909      0
#> 244  -0.080970161      0
#> 245  -0.064087804      0
#> 246   0.869927210      0
#> 247  -0.329252085      0
#> 248   0.198005201      0
#> 249  -0.464780519      0
#> 250  -1.113378878      0
#> 251  -2.494785292      0
#> 252   0.345366828      0
#> 253  -1.047141117      0
#> 254   0.814946228      0
#> 255   0.492669232      0
#> 256   0.638575451      0
#> 257  -0.867698978      0
#> 258   0.503458741      0
#> 259  -0.215569721      0
#> 260  -0.833379811      0
#> 261  -0.299344260      0
#> 262  -0.548994698      0
#> 263   1.183930619      0
#> 264   0.769621802      0
#> 265  -0.617680370      0
#> 266  -0.349300270      0
#> 267  -1.560956469      0
#> 268  -0.857648366      0
#> 269   0.011395309      0
#> 270  -1.649892561      0
#> 271   1.139659283      0
#> 272   1.660520564      0
#> 273   0.566986710      0
#> 274   1.227236590      0
#> 275  -0.114823321      0
#> 276   0.780587701      0
#> 277  -0.816608831      0
#> 278   0.528416005      0
#> 279   1.448470570      0
#> 280   0.060638677      0
#> 281   0.337560920      0
#> 282   2.255656946      0
#> 283   0.039434287      0
#> 284   0.603020387      0
#> 285   0.541537312      0
#> 286   0.488763503      0
#> 287   0.809669526      0
#> 288   1.144630825      0
#> 289   0.426833412      0
#> 290   0.556974612      0
#> 291   0.954697850      0
#> 292  -0.491983812      0
#> 293   1.565631415      0
#> 294   0.534713103      0
#> 295   0.092662604      0
#> 296   0.273505718      0
#> 297  -1.760321311      0
#> 298   2.201003393      0
#> 299   0.486111650      0
#> 300  -0.042839016      0
#> 301   0.024445260      0
#> 302  -0.535187592      0
#> 303  -1.670321445      0
#> 304  -1.613792368      0
#> 305  -1.732941657      0
#> 306   0.677368387      0
#> 307  -0.965007694      0
#> 308   0.444005768      0
#> 309  -1.375924758      0
#> 310   0.016858105      0
#> 311   1.098327550      0
#> 312   0.461475408      0
#> 313   0.898717101      0
#> 314  -0.990243169      0
#> 315  -0.241827198      0
#> 316  -0.951156671      0
#> 317   0.697196949      0
#> 318  -0.552227477      0
#> 319   1.068008513      0
#> 320   0.123563993      0
#> 321  -0.878671602      0
#> 322  -0.146215204      0
#> 323  -1.829020325      0
#> 324   1.505937576      0
#> 325  -0.573411849      0
#> 326   1.392119369      0
#> 327  -1.716567762      0
#> 328  -1.570044664      0
#> 329  -1.695623491      0
#> 330  -1.336065263      0
#> 331  -0.471680888      0
#> 332  -0.427095770      0
#> 333  -1.427100589      0
#> 334  -0.814313512      0
#> 335   0.111522166      0
#> 336  -1.165006427      0
#> 337  -1.219378949      0
#> 338   0.254437547      0
#> 339  -1.178796754      0
#> 340  -0.950612158      0
#> 341  -0.222195937      0
#> 342  -0.778512265      0
#> 343  -0.267597816      0
#> 344  -0.022985974      0
#> 345   0.750274584      0
#> 346   1.069640484      0
#> 347  -0.846482233      0
#> 348   0.252225707      0
#> 349  -1.077249415      0
#> 350  -0.274355539      0
#> 351  -1.212060842      0
#> 352  -0.120839839      0
#> 353  -0.269989164      0
#> 354   0.516863697      0
#> 355   0.004475778      0
#> 356  -1.761595497      0
#> 357  -0.484746634      0
#> 358  -0.291376774      0
#> 359  -0.050336955      0
#> 360  -0.921188035      0
#> 361   0.512964427      0
#> 362   0.379946436      0
#> 363  -0.039564542      0
#> 364   2.088414206      0
#> 365  -1.123115177      0
#> 366  -2.053565686      0
#> 367   0.906530500      0
#> 368   0.033120155      0
#> 369  -0.097361531      0
#> 370   0.237818076      0
#> 371  -0.224508750      0
#> 372   0.013658657      0
#> 373  -0.799726745      0
#> 374  -1.153569412      0
#> 375   0.680253144      0
#> 376   0.008324606      0
#> 377   0.399863016      0
#> 378   1.303770989      0
#> 379   0.805692580      0
#> 380  -0.873215657      0
#> 381  -0.744649411      0
#> 382  -2.095801554      0
#> 383   0.278162025      0
#> 384   0.268818969      0
#> 385  -0.489872180      0
#> 386   0.074822378      0
#> 387  -0.322117805      0
#> 388   0.966428498      0
#> 389   0.147570572      0
#> 390  -0.759765476      0
#> 391  -0.581781116      0
#> 392   0.854820449      0
#> 393   1.478955550      0
#> 394   1.193501761      0
#> 395  -0.126158789      0
#> 396   1.122519034      0
#> 397   0.807220316      0
#> 398   1.216649271      0
#> 399  -2.186628316      0
#> 400  -0.453929810      0
#> 401  -0.488235303      0
#> 402   1.698950555      0
#> 403   0.219789690      0
#> 404   0.319798752      0
#> 405  -2.644411485      0
#> 406   0.030652287      0
#> 407  -0.663798507      0
#> 408   0.127146076      0
#> 409   0.459486508      0
#> 410  -0.651238405      0
#> 411  -0.495966247      0
#> 412  -0.717942080      0
#> 413   1.606103689      0
#> 414   1.609925591      0
#> 415   1.596764957      0
#> 416  -1.100307154      0
#> 417   0.261553991      0
#> 418  -1.119589186      0
#> 419   0.019469790      0
#> 420  -0.359151150      0
#> 421   1.329573204      0
#> 422  -0.941000435      0
#> 423  -0.806817965      0
#> 424   0.348401676      0
#> 425   0.797660703      0
#> 426  -0.853105659      0
#> 427   0.143584721      0
#> 428   0.335213588      0
#> 429  -0.132698501      0
#> 430   1.472589944      0
#> 431  -0.014457739      0
#> 432   0.159666900      0
#> 433   0.208938179      0
#> 434   1.370905136      0
#> 435   0.397453945      0
#> 436   0.582205972      0
#> 437   2.297333905      0
#> 438  -2.890139366      0
#> 439  -0.336656793      0
#> 440   1.277993137      0
#> 441   2.647912488      0
#> 442   0.249929346      0
#> 443   0.364336723      0
#> 444   2.425326346      0
#> 445  -1.081917152      0
#> 446   0.430387116      0
#> 447  -0.688500214      0
#> 448   0.651456620      0
#> 449  -0.147795792      0
#> 450  -0.893295800      0
#> 451   0.312236166      0
#> 452   0.095397251      0
#> 453  -0.416667808      0
#> 454  -1.033193026      0
#> 455   1.039386279      0
#> 456   0.061600650      0
#> 457   0.882439261      0
#> 458   0.110592013      0
#> 459   0.874029260      0
#> 460  -0.196645521      0
#> 461   0.209837774      0
#> 462   1.183047205      0
#> 463   0.788010578      0
#> 464  -0.726648306      0
#> 465   0.427263786      0
#> 466   1.402081003      0
#> 467   0.903241948      0
#> 468   1.035417968      0
#> 469   0.291934830      0
#> 470  -0.594506206      0
#> 471   0.287386093      0
#> 472  -1.721004841      0
#> 473  -0.554680201      0
#> 474   0.338676686      0
#> 475   1.799798764      0
#> 476  -0.667457282      0
#> 477   0.729976490      0
#> 478   1.468964194      0
#> 479   1.031835255      0
#> 480   0.305406945      0
#> 481   1.733320351      0
#> 482   2.773108770      0
#> 483   0.207909738      0
#> 484   0.408196581      0
#> 485   0.785919337      0
#> 486   0.506956307      0
#> 487  -0.699366242      0
#> 488  -0.190077438      0
#> 489  -0.831602651      0
#> 490  -0.631782391      0
#> 491  -0.255539665      0
#> 492   0.041947851      0
#> 493   0.161550078      0
#> 494  -0.420741394      0
#> 495   0.002736740      0
#> 496   0.216822895      0
#> 497   1.451478389      0
#> 498  -0.365638158      0
#> 499   1.168739033      0
#> 500  -1.185463046      0
#> 501   1.199822525      0
#> 502  -0.951204240      0
#> 503   0.191184127      0
#> 504   2.018842579      0
#> 505  -0.110473702      0
#> 506  -0.639667370      0
#> 507  -1.220185567      0
#> 508  -1.273860826      0
#> 509  -0.565785611      0
#> 510   0.535450148      0
#> 511   0.847097689      0
#> 512  -0.471455290      0
#> 513  -1.324282643      0
#> 514  -0.411721785      0
#> 515   0.198971752      0
#> 516   0.199497924      0
#> 517   0.873509033      0
#> 518   0.586191668      0
#> 519   0.700104361      0
#> 520  -1.123834531      0
#> 521  -0.191169207      0
#> 522  -0.749008324      0
#> 523  -1.025443204      0
#> 524  -1.147488447      0
#> 525  -1.603692777      0
#> 526   0.128891562      0
#> 527   0.520502982      0
#> 528  -0.184890151      0
#> 529  -0.279650959      0
#> 530   2.665697479      0
#> 531   1.474738049      0
#> 532  -0.145937738      0
#> 533   0.374842626      0
#> 534  -0.734809501      0
#> 535  -1.834763771      0
#> 536  -0.005329923      0
#> 537  -0.030639902      0
#> 538   2.207963131      0
#> 539   1.226060488      0
#> 540   1.470082745      0
#> 541   0.347849406      0
#> 542   1.599332262      0
#> 543   2.084593768      0
#> 544   0.412321002      0
#> 545  -1.418639231      0
#> 546   0.102322053      0
#> 547   0.325624176      0
#> 548   1.606162941      0
#> 549   1.093602274      0
#> 550  -0.168804108      0
#> 551  -2.252850172      0
#> 552   0.408641590      0
#> 553  -1.201665725      0
#> 554   0.360751597      0
#> 555   0.774037223      0
#> 556  -0.589682858      0
#> 557   0.878793293      0
#> 558  -0.526316701      0
#> 559  -0.497287602      0
#> 560  -1.152754419      0
#> 561  -0.807923905      0
#> 562  -0.058997242      0
#> 563   0.133445096      0
#> 564   0.436364977      0
#> 565   0.217644669      0
#> 566  -1.142442879      0
#> 567   1.487102701      0
#> 568  -1.069893741      0
#> 569   0.541126916      0
#> 570  -0.882831384      0
#> 571  -1.783116857      0
#> 572  -1.345708270      0
#> 573  -2.329955696      0
#> 574  -1.760450479      0
#> 575   0.245568170      0
#> 576  -0.187181730      0
#> 577   0.155258974      0
#> 578  -0.887665932      0
#> 579   0.706729363      0
#> 580  -1.405525598      0
#> 581   0.527617988      0
#> 582   0.098623760      0
#> 583  -0.487147010      0
#> 584  -1.461565788      0
#> 585   0.655595157      0
#> 586   0.855826172      0
#> 587   0.486811147      0
#> 588   2.010584965      0
#> 589  -1.582022201      0
#> 590   1.370383392      0
#> 591  -0.249849935      0
#> 592   0.572625915      0
#> 593  -0.081718569      0
#> 594  -0.608169531      0
#> 595  -1.253842656      0
#> 596  -0.530201941      0
#> 597  -1.160544015      0
#> 598  -2.220345138      0
#> 599  -1.142708163      0
#> 600   0.870191293      0
#> 601   0.848028368      0
#> 602  -0.049504529      0
#> 603   0.319238896      0
#> 604  -0.414319741      0
#> 605   0.917618731      0
#> 606  -1.190393155      0
#> 607  -0.051122683      0
#> 608   0.238018028      0
#> 609  -2.045870247      0
#> 610  -2.237520445      0
#> 611  -0.689612785      0
#> 612   0.920403446      0
#> 613   0.298654990      0
#> 614   1.251302565      0
#> 615  -1.067917742      0
#> 616   0.942372321      0
#> 617  -0.584102392      0
#> 618   0.581063283      0
#> 619   0.377806135      0
#> 620  -0.096300155      0
#> 621   2.062424809      0
#> 622  -0.086620266      0
#> 623   0.550957696      0
#> 624  -0.050922142      0
#> 625   0.811474638      0
#> 626   0.822181093      0
#> 627   1.004893645      0
#> 628   0.489933164      0
#> 629  -1.041102664      0
#> 630  -0.480531421      0
#> 631   0.764694272      0
#> 632   0.739630002      0
#> 633   1.780828188      0
#> 634   0.183206563      0
#> 635  -0.564541131      0
#> 636   0.796468633      0
#> 637   0.843992195      0
#> 638   0.161154474      0
#> 639   0.608937813      0
#> 640   0.212412195      0
#> 641  -1.335999061      0
#> 642   0.296012322      0
#> 643  -1.099688987      0
#> 644   0.093937145      0
#> 645   0.307612681      0
#> 646   0.092705649      0
#> 647   0.539355365      0
#> 648  -0.994548022      0
#> 649   1.460406738      0
#> 650   0.156571422      0
#> 651   1.435899168      0
#> 652  -0.964622042      0
#> 653   0.156094409      0
#> 654  -1.314044252      0
#> 655  -0.017709863      0
#> 656   0.666303251      0
#> 657   0.514161434      0
#> 658   0.098036319      0
#> 659   0.895042132      0
#> 660  -1.780945218      0
#> 661  -0.065372001      0
#> 662  -0.298847321      0
#> 663  -0.383989709      0
#> 664   2.454020099      0
#> 665   1.782468404      0
#> 666  -1.430463673      0
#> 667   1.014730839      0
#> 668   0.832688434      0
#> 669  -1.940748701      0
#> 670   0.058255222      0
#> 671   1.208063225      0
#> 672   1.507898322      0
#> 673  -0.456480234      0
#> 674   0.721604537      0
#> 675   0.437416718      0
#> 676  -0.399426371      0
#> 677   0.286071257      0
#> 678  -0.515336940      0
#> 679   1.338759852      0
#> 680   0.741462850      0
#> 681   0.455934903      0
#> 682  -0.251997826      0
#> 683  -1.031543810      0
#> 684   0.753029721      0
#> 685   1.285917776      0
#> 686  -1.231549676      0
#> 687   0.559322230      0
#> 688   0.265720163      0
#> 689  -0.082285941      0
#> 690  -0.217700423      0
#> 691  -0.082711738      0
#> 692  -0.693945172      0
#> 693   0.730358371      0
#> 694   0.329988351      0
#> 695  -0.154960032      0
#> 696   0.367997239      0
#> 697   0.948792780      0
#> 698   1.131685429      0
#> 699   0.005107499      0
#> 700   0.908913502      0
#> 701   0.245908980      0
#> 702   1.200657939      0
#> 703   2.461471589      0
#> 704  -0.187668626      0
#> 705  -0.071734926      0
#> 706  -1.925152619      0
#> 707   2.315310098      0
#> 708  -0.626974923      0
#> 709  -0.186804052      0
#> 710   0.148645348      0
#> 711   0.439628663      0
#> 712   0.309790047      0
#> 713  -1.410975418      0
#> 714   0.761817898      0
#> 715  -1.015614791      0
#> 716   0.384975187      0
#> 717   0.142054908      0
#> 718   1.041512012      0
#> 719   0.799121373      0
#> 720   0.496315821      0
#> 721   0.381237589      0
#> 722   1.055228579      0
#> 723  -1.215772123      0
#> 724  -0.601798092      0
#> 725   0.889567026      0
#> 726  -0.517184781      0
#> 727   1.130207658      0
#> 728  -0.924595141      0
#> 729   0.877750952      0
#> 730  -1.846468871      0
#> 731   1.254651664      0
#> 732   0.322576725      0
#> 733  -0.583476568      0
#> 734  -1.768690146      0
#> 735   0.523265031      0
#> 736  -0.129730549      0
#> 737   0.884897721      0
#> 738   0.089548866      0
#> 739  -0.883555950      0
#> 740  -0.429810845      0
#> 741  -0.352122910      0
#> 742   0.009304653      0
#> 743   0.938290678      0
#> 744  -0.154174267      0
#> 745   1.279434116      0
#> 746   1.000668817      0
#> 747   0.543754828      0
#> 748   0.761680312      0
#> 749   0.256243467      0
#> 750   0.651536736      0
#> 751  -0.620366142      0
#> 752   1.641625701      0
#> 753  -1.412637059      0
#> 754   0.120352394      0
#> 755   1.182367016      0
#> 756   0.220473026      0
#> 757   0.986802391      0
#> 758   1.966909130      0
#> 759   0.910617997      0
#> 760  -0.519984799      0
#> 761   0.021549693      0
#> 762  -1.060515855      0
#> 763   2.044481437      0
#> 764   1.186774953      0
#> 765   0.039595187      0
#> 766  -0.182772991      0
#> 767  -1.162942601      0
#> 768   0.729079108      0
#> 769   0.603138217      0
#> 770   0.060224036      0
#> 771  -0.626184747      0
#> 772   0.809000099      0
#> 773  -0.435812136      0
#> 774   1.320791337      0
#> 775   0.484974988      0
#> 776   0.736572283      0
#> 777   1.160562964      0
#> 778  -0.163791522      0
#> 779  -0.191645829      0
#> 780   2.548447536      0
#> 781   1.228845964      0
#> 782   0.255304369      0
#> 783   0.036420572      0
#> 784   1.664566616      0
#> 785  -0.534542521      0
#> 786  -2.813017104      0
#> 787  -1.698954743      0
#> 788   0.713665208      0
#> 789   0.912852399      0
#> 790  -1.245131849      0
#> 791   0.118144666      0
#> 792   0.885373737      0
#> 793   0.167168662      0
#> 794   0.117167820      0
#> 795   0.552461161      0
#> 796  -0.555290346      0
#> 797  -1.218910810      0
#> 798   0.765509214      0
#> 799   0.450408170      0
#> 800  -2.272378117      0
#> 801  -0.443919546      0
#> 802  -2.594936624      0
#> 803  -0.167248203      0
#> 804  -0.873872039      0
#> 805  -0.771321272      0
#> 806   0.124288185      0
#> 807  -0.581733116      0
#> 808  -0.342282084      0
#> 809  -1.234018621      0
#> 810  -0.983145241      0
#> 811  -0.508302032      0
#> 812  -0.583489330      0
#> 813  -1.376752876      0
#> 814   0.174038268      0
#> 815   0.582143896      0
#> 816   1.042101807      0
#> 817   0.057589379      0
#> 818   0.227058708      0
#> 819  -0.160707175      0
#> 820   0.486911801      0
#> 821  -0.056863894      0
#> 822  -1.427926248      0
#> 823  -0.753638038      0
#> 824  -1.129639363      0
#> 825  -1.336482133      0
#> 826  -1.777990988      0
#> 827   0.573543624      0
#> 828   0.120906615      0
#> 829   0.357235425      0
#> 830  -0.152039186      0
#> 831   0.499826612      0
#> 832  -0.044676729      0
#> 833  -0.004720621      0
#> 834   0.145440046      0
#> 835  -0.303913794      0
#> 836  -0.009373467      0
#> 837   0.044081779      0
#> 838  -0.196580262      0
#> 839  -1.521441521      0
#> 840   1.371069584      0
#> 841   0.115658287      0
#> 842   2.122531549      0
#> 843  -1.062563722      0
#> 844  -0.099720206      0
#> 845   0.166899715      0
#> 846  -1.757179890      0
#> 847  -0.318049077      0
#> 848  -0.239329720      0
#> 849   0.433382660      0
#> 850  -0.322633041      0
#> 851  -0.278592708      0
#> 852  -1.266622600      0
#> 853   0.025374223      0
#> 854  -0.122658811      0
#> 855   0.729749827      0
#> 856   1.490422450      0
#> 857  -2.189321121      0
#> 858  -0.161001704      0
#> 859  -0.653965118      0
#> 860   0.996576734      0
#> 861   1.441845680      0
#> 862  -1.342070320      0
#> 863   0.467252678      0
#> 864  -0.928474793      0
#> 865   0.308592494      0
#> 866  -0.091939492      0
#> 867  -0.867435733      0
#> 868  -2.022174623      0
#> 869   1.649841256      0
#> 870   0.434442582      0
#> 871   0.278711278      0
#> 872   0.206914292      0
#> 873  -0.858640603      0
#> 874   1.775541804      0
#> 875   0.391261084      0
#> 876  -1.498001961      0
#> 877  -0.087269391      0
#> 878  -1.004645524      0
#> 879   0.200237497      0
#> 880   1.230558495      0
#> 881  -0.541120082      0
#> 882   1.165910420      0
#> 883  -0.031403799      0
#> 884   1.500260678      0
#> 885   0.911869107      0
#> 886   0.903120010      0
#> 887  -0.079514396      0
#> 888  -1.340233301      0
#> 889   0.423127502      0
#> 890   0.415547018      0
#> 891  -0.082432221      0
#> 892   2.234186913      1
#> 893   1.090011006      1
#> 894   2.037872005      1
#> 895  -0.043500896      1
#> 896   0.880479554      1
#> 897   0.469945031      1
#> 898   0.748542982      1
#> 899  -0.161877962      1
#> 900   2.073402899      1
#> 901   1.367462506      1
#> 902  -0.460626827      1
#> 903  -0.145185105      1
#> 904   2.414842375      1
#> 905   1.755300280      1
#> 906   2.828818519      1
#> 907   2.317393560      1
#> 908   0.695509509      1
#> 909   0.950357959      1
#> 910   1.385138679      1
#> 911   0.796996380      1
#> 912  -0.197003411      1
#> 913   1.826414195      1
#> 914   0.852257591      1
#> 915  -0.187468820      1
#> 916   1.361119983      1
#> 917   1.967472586      1
#> 918  -0.633186666      1
#> 919   1.010336310      1
#> 920  -0.302790526      1
#> 921  -0.676232786      1
#> 922   2.792993564      1
#> 923   1.463787029      1
#> 924   0.778127826      1
#> 925   0.065622780      1
#> 926   1.699721619      1
#> 927   0.134078963      1
#> 928   3.278954514      1
#> 929   1.868667847      1
#> 930   1.617548467      1
#> 931   1.194368475      1
#> 932   1.253092602      1
#> 933   0.758855179      1
#> 934   2.532946960      1
#> 935   0.275029644      1
#> 936   0.885987616      1
#> 937   0.508670067      1
#> 938   0.557709963      1
#> 939  -0.065224586      1
#> 940   2.412857091      1
#> 941   0.696581720      1
#> 942   1.976960910      1
#> 943   1.199776697      1
#> 944   0.245541959      1
#> 945   0.425057932      1
#> 946  -0.070385468      1
#> 947   1.034079791      1
#> 948   1.928205494      1
#> 949  -1.600802919      1
#> 950   0.492064814      1
#> 951   0.358572195      1
#> 952   0.709924508      1
#> 953  -2.352425269      1
#> 954   1.068278954      1
#> 955   0.203333071      1
#> 956   1.341752504      1
#> 957   0.441937578      1
#> 958   1.922892844      1
#> 959   1.639842081      1
#> 960   0.463565318      1
#> 961   0.858284619      1
#> 962   1.426018980      1
#> 963  -0.343113274      1
#> 964   0.517686180      1
#> 965   0.574614861      1
#> 966   0.391241000      1
#> 967   1.409978902      1
#> 968   0.370497477      1
#> 969   1.712461268      1
#> 970   2.131781677      1
#> 971  -0.508730204      1
#> 972   0.425929278      1
#> 973  -0.055679025      1
#> 974   0.641846395      1
#> 975  -0.391059626      1
#> 976  -0.627698894      1
#> 977   2.259760917      1
#> 978   0.353709277      1
#> 979  -0.288083801      1
#> 980   1.194332720      1
#> 981   0.324299462      1
#> 982  -0.132090028      1
#> 983  -0.349312566      1
#> 984   1.345584985      1
#> 985   1.145959259      1
#> 986  -0.053092308      1
#> 987   2.885641486      1
#> 988   0.463780465      1
#> 989   0.812032569      1
#> 990   0.045365331      1
#> 991  -0.432122405      1
#> 992   1.399440912      1
#> 993   0.685318441      1
#> 994  -0.143380735      1
#> 995  -0.262400521      1
#> 996   0.705103017      1
#> 997   0.914603318      1
#> 998   1.449068486      1
#> 999  -0.740415029      1
#> 1000 -0.419810612      1