Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.312668903      0
#> 2    -0.336993776      0
#> 3     0.275709912      0
#> 4    -0.307839852      0
#> 5    -0.442045111      0
#> 6     0.648878665      0
#> 7     1.146800873      0
#> 8    -0.437217730      0
#> 9    -0.700796703      0
#> 10   -0.892112412      0
#> 11   -1.603765818      0
#> 12   -0.927964386      0
#> 13   -0.148879610      0
#> 14   -0.043160607      0
#> 15    0.514156236      0
#> 16    0.502186273      0
#> 17   -1.344359938      0
#> 18    0.721028806      0
#> 19   -0.386765580      0
#> 20    0.349268138      0
#> 21   -0.621976009      0
#> 22   -0.103807083      0
#> 23   -0.488950699      0
#> 24    0.136247935      0
#> 25    0.589133037      0
#> 26   -0.491983757      0
#> 27   -0.149377587      0
#> 28    1.040129066      0
#> 29    0.307558686      0
#> 30    1.427981551      0
#> 31    0.576857765      0
#> 32   -0.648871107      0
#> 33   -0.250259198      0
#> 34    0.455327470      0
#> 35    0.183009473      0
#> 36    1.247074017      0
#> 37    1.432871288      0
#> 38    0.582097683      0
#> 39    0.094675052      0
#> 40   -0.470985903      0
#> 41   -0.211652404      0
#> 42   -1.835952477      0
#> 43    0.642111411      0
#> 44    2.422039078      0
#> 45   -1.767299038      0
#> 46    0.485967174      0
#> 47    1.802569421      0
#> 48   -0.142136604      0
#> 49   -1.093162640      0
#> 50    0.642807612      0
#> 51    0.234305726      0
#> 52   -0.736303160      0
#> 53    0.266275926      0
#> 54   -0.591228267      0
#> 55    1.754847743      0
#> 56    0.012795017      0
#> 57    1.114457606      0
#> 58   -0.470655142      0
#> 59    0.701911843      0
#> 60   -1.073428068      0
#> 61   -0.837517247      0
#> 62    1.339877138      0
#> 63    0.257107929      0
#> 64    0.344497262      0
#> 65    1.096372999      0
#> 66   -0.598355206      0
#> 67    0.232036556      0
#> 68   -1.446245530      0
#> 69   -1.170336399      0
#> 70   -0.587777054      0
#> 71    0.061777548      0
#> 72    1.610181271      0
#> 73   -1.220144900      0
#> 74    1.581310507      0
#> 75    0.203621062      0
#> 76    0.734295481      0
#> 77    0.364880871      0
#> 78   -0.089104818      0
#> 79    0.681978053      0
#> 80   -1.099485779      0
#> 81   -0.231080218      0
#> 82    0.452588458      0
#> 83    0.111325850      0
#> 84    0.401589029      0
#> 85   -0.417005405      0
#> 86   -0.050139688      0
#> 87    0.477094143      0
#> 88    0.767188631      0
#> 89   -0.200202626      0
#> 90   -0.975076360      0
#> 91   -0.141030015      0
#> 92    0.125755977      0
#> 93    2.423845075      0
#> 94    0.044146735      0
#> 95    0.420550607      0
#> 96    0.450696952      0
#> 97   -2.603239147      0
#> 98    0.240386281      0
#> 99    0.276333777      0
#> 100   0.015648471      0
#> 101  -0.894810128      0
#> 102   1.214583345      0
#> 103   1.174835682      0
#> 104   0.535495500      0
#> 105   0.980330775      0
#> 106  -0.181236193      0
#> 107  -0.914354087      0
#> 108   1.131432751      0
#> 109   2.659453028      0
#> 110   1.835971668      0
#> 111  -0.171297677      0
#> 112   1.350209494      0
#> 113   1.155942461      0
#> 114   0.387169010      0
#> 115  -0.673452309      0
#> 116   1.152272161      0
#> 117  -0.439957390      0
#> 118   0.205126275      0
#> 119  -0.258380427      0
#> 120  -0.776852885      0
#> 121  -1.052797562      0
#> 122   0.407397744      0
#> 123  -1.422707762      0
#> 124  -1.425219480      0
#> 125  -0.017520467      0
#> 126  -0.479281815      0
#> 127  -2.617458969      0
#> 128   0.372018643      0
#> 129  -0.933339210      0
#> 130  -1.123583670      0
#> 131  -1.101210603      0
#> 132  -0.639574576      0
#> 133  -0.965877833      0
#> 134  -1.305714133      0
#> 135  -0.077989497      0
#> 136  -1.235576581      0
#> 137  -1.233227751      0
#> 138   0.116337878      0
#> 139   0.836449153      0
#> 140  -0.395113783      0
#> 141  -0.053782007      0
#> 142  -1.374011503      0
#> 143   0.044890441      0
#> 144  -2.971702509      0
#> 145  -1.025216402      0
#> 146  -0.400696004      0
#> 147  -0.845112202      0
#> 148  -0.319730870      0
#> 149  -0.847172357      0
#> 150  -1.195769488      0
#> 151  -0.002873980      0
#> 152  -1.801071503      0
#> 153  -1.333834452      0
#> 154  -0.681580764      0
#> 155   1.753002939      0
#> 156   0.238183154      0
#> 157   0.586955181      0
#> 158  -0.649176459      0
#> 159  -1.564664137      0
#> 160  -1.956485920      0
#> 161  -1.147884640      0
#> 162  -0.316221777      0
#> 163   0.220647052      0
#> 164   1.002368025      0
#> 165   0.383101001      0
#> 166  -0.735656382      0
#> 167   0.151053852      0
#> 168  -0.172917500      0
#> 169   0.424624838      0
#> 170   0.918087099      0
#> 171   1.345530354      0
#> 172  -0.015857795      0
#> 173   0.238545702      0
#> 174  -0.480043990      0
#> 175  -0.927542344      0
#> 176  -0.814673066      0
#> 177  -0.951508492      0
#> 178  -0.160785148      0
#> 179  -0.208230454      0
#> 180   0.628435179      0
#> 181  -0.319086847      0
#> 182  -0.687951486      0
#> 183  -1.960414353      0
#> 184   1.446095171      0
#> 185  -0.819094283      0
#> 186   1.325689021      0
#> 187  -0.387015255      0
#> 188   0.328187480      0
#> 189  -0.059663040      0
#> 190   0.210437664      0
#> 191  -0.215202130      0
#> 192   0.065718553      0
#> 193  -0.680591154      0
#> 194   1.670693231      0
#> 195  -0.397820395      0
#> 196  -0.562061309      0
#> 197  -2.019442065      0
#> 198   0.589708264      0
#> 199   0.096178161      0
#> 200  -0.125217110      0
#> 201  -0.527728926      0
#> 202   0.481796041      0
#> 203  -0.603822001      0
#> 204  -0.076922471      0
#> 205   0.538299067      0
#> 206  -0.319459869      0
#> 207  -1.473934112      0
#> 208  -0.410924749      0
#> 209  -0.287805598      0
#> 210  -0.573270184      0
#> 211  -1.135185787      0
#> 212   1.174942480      0
#> 213   0.226655404      0
#> 214   0.735720162      0
#> 215   0.230777483      0
#> 216   0.249750095      0
#> 217  -0.327437376      0
#> 218  -0.071374420      0
#> 219  -1.770478368      0
#> 220   1.353840835      0
#> 221  -0.367741936      0
#> 222  -1.366702730      0
#> 223  -0.804125439      0
#> 224   1.069284784      0
#> 225  -0.063642385      0
#> 226  -0.142465031      0
#> 227   0.232447338      0
#> 228  -2.330972140      0
#> 229   1.537548204      0
#> 230  -2.345941019      0
#> 231  -0.813187591      0
#> 232  -0.710700308      0
#> 233   0.618969254      0
#> 234   0.488151607      0
#> 235   0.745172890      0
#> 236   0.929922740      0
#> 237   0.962143641      0
#> 238   1.423865094      0
#> 239  -1.352663019      0
#> 240   1.641408560      0
#> 241   1.352520421      0
#> 242   0.126407695      0
#> 243   0.372548479      0
#> 244   1.199909297      0
#> 245   0.083440503      0
#> 246  -0.484467567      0
#> 247  -0.786216090      0
#> 248  -1.575362901      0
#> 249  -0.420588584      0
#> 250   0.692608269      0
#> 251   0.046919360      0
#> 252  -0.224755815      0
#> 253   0.477464152      0
#> 254   0.404988774      0
#> 255   0.089532946      0
#> 256  -0.088280083      0
#> 257  -0.685674704      0
#> 258  -1.096604855      0
#> 259   1.594395633      0
#> 260  -1.877566076      0
#> 261   0.747778079      0
#> 262   0.243996925      0
#> 263   1.414591211      0
#> 264   1.965105983      0
#> 265   0.326089913      0
#> 266  -0.248542183      0
#> 267   0.520395923      0
#> 268   1.266173945      0
#> 269  -0.961945226      0
#> 270   0.419620351      0
#> 271  -0.403755229      0
#> 272   0.512876204      0
#> 273  -1.735050971      0
#> 274   0.806977156      0
#> 275   0.800317179      0
#> 276  -0.770136763      0
#> 277   1.639992008      0
#> 278   0.468203557      0
#> 279   0.735571407      0
#> 280   1.103699440      0
#> 281   0.252997779      0
#> 282   1.143625302      0
#> 283   1.638832840      0
#> 284  -0.123181883      0
#> 285  -0.046799378      0
#> 286   1.016632691      0
#> 287   1.310445352      0
#> 288  -0.786026920      0
#> 289  -2.322636485      0
#> 290   0.822501740      0
#> 291  -0.167350135      0
#> 292  -1.162532020      0
#> 293  -0.179382613      0
#> 294  -0.682520843      0
#> 295  -0.050588959      0
#> 296  -0.366313400      0
#> 297   0.790501697      0
#> 298  -0.470447474      0
#> 299   2.109474187      0
#> 300   0.664077745      0
#> 301   1.901859697      0
#> 302  -1.307057778      0
#> 303  -0.860237638      0
#> 304   1.580202439      0
#> 305  -1.066645230      0
#> 306   0.076641744      0
#> 307   1.898041474      0
#> 308  -0.469413054      0
#> 309   1.773957848      0
#> 310  -0.021959611      0
#> 311  -0.044471950      0
#> 312   0.733700980      0
#> 313   0.379323897      0
#> 314   1.043081706      0
#> 315   1.137284311      0
#> 316   0.860518551      0
#> 317  -2.522941429      0
#> 318   0.190082582      0
#> 319  -0.009689414      0
#> 320   0.280134490      0
#> 321   2.038931863      0
#> 322  -1.735202607      0
#> 323  -1.506162051      0
#> 324   0.469480267      0
#> 325   0.711407516      0
#> 326   0.675168025      0
#> 327   0.089290103      0
#> 328  -1.075746665      0
#> 329   0.162101742      0
#> 330  -1.028562497      0
#> 331  -0.263730441      0
#> 332   0.658897166      0
#> 333  -0.288659832      0
#> 334   0.882522563      0
#> 335   0.084171955      0
#> 336  -0.326487013      0
#> 337  -1.716158208      0
#> 338  -0.234091369      0
#> 339   1.061118096      0
#> 340   0.597121721      0
#> 341  -1.035163315      0
#> 342   0.154552865      0
#> 343  -0.302863599      0
#> 344  -1.472669944      0
#> 345   0.660647696      0
#> 346  -0.144880752      0
#> 347  -1.126670627      0
#> 348   0.345964447      0
#> 349   1.116975193      0
#> 350   0.156542762      0
#> 351   0.949740027      0
#> 352   0.983730325      0
#> 353  -0.444512376      0
#> 354  -1.370481285      0
#> 355  -1.222895190      0
#> 356  -0.646309687      0
#> 357  -0.647091440      0
#> 358   1.323529612      0
#> 359   1.167848997      0
#> 360  -0.812799829      0
#> 361  -0.009777656      0
#> 362  -0.632746514      0
#> 363   0.607305314      0
#> 364   1.127866730      0
#> 365   0.401697990      0
#> 366  -0.062045340      0
#> 367   0.315142111      0
#> 368  -1.019349158      0
#> 369  -0.525730311      0
#> 370   1.256199429      0
#> 371  -0.030031831      0
#> 372  -1.189932179      0
#> 373  -0.061041284      0
#> 374  -0.460087939      0
#> 375   1.735030505      0
#> 376   0.231264459      0
#> 377   0.018831045      0
#> 378   0.221993288      0
#> 379  -1.206562474      0
#> 380   1.371244663      0
#> 381   0.390595080      0
#> 382   0.311931071      0
#> 383   1.100703000      0
#> 384  -0.771858593      0
#> 385   0.683621129      0
#> 386   0.006013099      0
#> 387   0.259989878      0
#> 388  -0.434824573      0
#> 389  -1.655205260      0
#> 390   0.852744745      0
#> 391  -1.495814092      0
#> 392   0.881447451      0
#> 393   2.422962851      0
#> 394   0.758630798      0
#> 395   0.433579148      0
#> 396   0.642095381      0
#> 397  -0.565163709      0
#> 398  -1.306104655      0
#> 399  -1.442123212      0
#> 400   0.789310567      0
#> 401  -0.711095870      0
#> 402   0.994207152      0
#> 403  -1.548869024      0
#> 404  -1.029340634      0
#> 405  -0.014336018      0
#> 406   1.406276923      0
#> 407  -1.928341072      0
#> 408  -1.647016764      0
#> 409  -0.417516257      0
#> 410  -0.113243466      0
#> 411   0.272761615      0
#> 412  -1.810277645      0
#> 413   0.546735163      0
#> 414  -0.154846082      0
#> 415  -0.724589421      0
#> 416   0.403524935      0
#> 417  -0.819819165      0
#> 418  -1.306138211      0
#> 419   0.297725363      0
#> 420  -0.528504682      0
#> 421   0.496430383      0
#> 422   0.284448042      0
#> 423   0.500912865      0
#> 424  -1.175630329      0
#> 425  -0.917099370      0
#> 426   0.109251133      0
#> 427   0.857501610      0
#> 428   1.662583970      0
#> 429   0.364295188      0
#> 430  -0.359900553      0
#> 431   0.407034743      0
#> 432  -0.072295519      0
#> 433   0.161664015      0
#> 434   0.710629958      0
#> 435  -1.530511571      0
#> 436   1.077410961      0
#> 437   0.670497891      0
#> 438  -1.674920316      0
#> 439   1.977915086      0
#> 440  -1.103098549      0
#> 441  -0.143321611      0
#> 442  -1.118388948      0
#> 443   0.371155403      0
#> 444   2.235132100      0
#> 445  -0.373067466      0
#> 446   0.123355570      0
#> 447  -0.497414828      0
#> 448   0.279069460      0
#> 449  -0.822992478      0
#> 450  -1.590182266      0
#> 451  -0.597519010      0
#> 452  -1.123863487      0
#> 453   1.073122316      0
#> 454  -0.131975695      0
#> 455   2.332172320      0
#> 456  -0.968689195      0
#> 457  -0.555620784      0
#> 458  -0.490464693      0
#> 459  -0.103044415      0
#> 460   1.923023840      0
#> 461   0.756214801      0
#> 462  -1.158432834      0
#> 463  -0.554390858      0
#> 464  -1.208002561      0
#> 465  -1.054378611      0
#> 466  -0.506775373      0
#> 467   0.733870803      0
#> 468  -0.563417039      0
#> 469  -0.479719769      0
#> 470  -0.582942406      0
#> 471  -0.546514797      0
#> 472   0.021028705      0
#> 473   0.487386141      0
#> 474  -0.393092553      0
#> 475  -0.430038478      0
#> 476  -1.465339848      0
#> 477   0.604579096      0
#> 478  -0.460627299      0
#> 479  -0.492541765      0
#> 480   0.889378723      0
#> 481  -0.362276918      0
#> 482   0.377157867      0
#> 483  -0.520674962      0
#> 484  -0.914438169      0
#> 485  -0.298112732      0
#> 486  -0.260925943      0
#> 487  -0.868566509      0
#> 488  -0.430269340      0
#> 489   1.367016697      0
#> 490   0.023662294      0
#> 491  -0.507976119      0
#> 492   0.099677168      0
#> 493   0.721689330      0
#> 494   1.036373277      0
#> 495   0.776445479      0
#> 496  -1.211586596      0
#> 497  -0.049874688      0
#> 498  -0.117649287      0
#> 499   1.018606364      0
#> 500  -0.084323264      0
#> 501   0.717694915      0
#> 502   0.939971315      0
#> 503  -0.285094337      0
#> 504   0.368479072      0
#> 505  -0.277243901      0
#> 506   1.209725587      0
#> 507   0.749669054      0
#> 508  -0.826642877      0
#> 509   0.895270199      0
#> 510  -0.976351976      0
#> 511  -0.618084238      0
#> 512  -1.489022287      0
#> 513   1.961827948      0
#> 514   1.912363950      0
#> 515  -0.047135846      0
#> 516  -0.155190313      0
#> 517  -0.041852330      0
#> 518   1.069924935      0
#> 519  -0.795798556      0
#> 520   0.364047592      0
#> 521  -0.029615278      0
#> 522  -0.815516139      0
#> 523  -0.865010804      0
#> 524   1.251288852      0
#> 525   0.251921223      0
#> 526  -0.839176702      0
#> 527  -0.035382423      0
#> 528  -1.041560616      0
#> 529   0.992990267      0
#> 530   0.959214102      0
#> 531  -0.554511955      0
#> 532   0.572310236      0
#> 533  -0.317216116      0
#> 534   1.085482840      0
#> 535   1.577965341      0
#> 536  -0.816947794      0
#> 537  -0.987013424      0
#> 538   0.976071180      0
#> 539  -0.581577500      0
#> 540  -2.773719114      0
#> 541   0.936769195      0
#> 542   0.371476478      0
#> 543  -0.923949079      0
#> 544   0.112588924      0
#> 545   0.932432479      0
#> 546   1.476017965      0
#> 547  -1.187065745      0
#> 548  -0.054778231      0
#> 549  -0.963758405      0
#> 550  -0.412028242      0
#> 551  -0.935088265      0
#> 552   0.786217708      0
#> 553  -0.928041749      0
#> 554  -0.079289073      0
#> 555  -0.321346606      0
#> 556  -1.303518969      0
#> 557   0.909881695      0
#> 558   1.389931201      0
#> 559  -1.036005533      0
#> 560  -2.838232830      0
#> 561   0.893211460      0
#> 562   0.219995360      0
#> 563   0.770327072      0
#> 564  -0.383761135      0
#> 565  -0.139551320      0
#> 566  -1.062162654      0
#> 567   1.296162126      0
#> 568   0.085218525      0
#> 569   1.298301330      0
#> 570   0.017382760      0
#> 571   0.170377723      0
#> 572  -0.995128292      0
#> 573   0.855769455      0
#> 574  -0.235460199      0
#> 575  -1.156911004      0
#> 576  -1.167720639      0
#> 577  -0.824517122      0
#> 578   0.059913708      0
#> 579   0.056825859      0
#> 580  -0.364797286      0
#> 581   1.099244153      0
#> 582   0.095405159      0
#> 583   0.833180900      0
#> 584   1.760429880      0
#> 585  -1.602342345      0
#> 586  -1.450470991      0
#> 587  -0.273267728      0
#> 588  -1.337159219      0
#> 589   1.704485074      0
#> 590   1.546834098      0
#> 591  -0.561143510      0
#> 592   1.663155501      0
#> 593  -0.950339587      0
#> 594  -0.017420077      0
#> 595   0.453662007      0
#> 596  -0.725892302      0
#> 597  -0.327059851      0
#> 598  -0.557322561      0
#> 599  -0.345493546      0
#> 600  -0.603348022      0
#> 601  -0.495862236      0
#> 602   0.687161191      0
#> 603  -0.487407796      0
#> 604   0.042255353      0
#> 605   0.143407071      0
#> 606   1.040819253      0
#> 607   0.568953743      0
#> 608  -0.175682093      0
#> 609   0.767798485      0
#> 610   1.521194249      0
#> 611  -1.239400706      0
#> 612   0.008178102      0
#> 613   0.757221372      0
#> 614   1.565364744      0
#> 615  -0.393377660      0
#> 616  -0.031594663      0
#> 617  -0.100840127      0
#> 618  -1.241195743      0
#> 619   0.547884329      0
#> 620  -0.768545157      0
#> 621   0.335791438      0
#> 622  -0.967985771      0
#> 623  -1.978300289      0
#> 624  -0.782445864      0
#> 625  -1.239909950      0
#> 626   0.752868258      0
#> 627  -0.355907928      0
#> 628   0.931048128      0
#> 629  -1.005584120      0
#> 630   2.224069863      0
#> 631   1.158858190      0
#> 632   0.697722416      0
#> 633   0.868032576      0
#> 634   0.338096227      0
#> 635   2.154489954      0
#> 636  -0.847068545      0
#> 637   0.280992694      0
#> 638  -1.310337689      0
#> 639  -2.823678548      0
#> 640   0.388259098      0
#> 641  -0.311500779      0
#> 642   0.197602040      0
#> 643  -0.237670966      0
#> 644   0.372180556      0
#> 645  -1.471743118      0
#> 646  -1.498318645      0
#> 647   0.537626073      0
#> 648  -1.167505337      0
#> 649   0.022957157      0
#> 650  -0.799275160      0
#> 651   1.100363272      0
#> 652  -0.318411590      0
#> 653   0.398357215      0
#> 654   0.027162071      0
#> 655  -0.350190076      0
#> 656  -0.575644110      0
#> 657  -1.968301190      0
#> 658  -0.618048737      0
#> 659   0.173189322      0
#> 660   0.060568875      0
#> 661   0.094559870      0
#> 662  -0.127788774      0
#> 663  -0.017672685      0
#> 664  -0.425859562      0
#> 665   1.186552120      0
#> 666  -1.587609278      0
#> 667  -0.848734452      0
#> 668   0.163556227      0
#> 669  -0.022235874      0
#> 670  -0.598142229      0
#> 671   0.002166245      0
#> 672  -0.278808652      0
#> 673   1.357181674      0
#> 674  -0.957157323      0
#> 675   1.812379975      0
#> 676   0.329332947      0
#> 677  -0.160711989      0
#> 678  -2.977246935      0
#> 679  -0.814542948      0
#> 680   0.120060068      0
#> 681   0.228510340      0
#> 682   1.142750977      0
#> 683   0.700755695      0
#> 684  -0.242034860      0
#> 685  -0.341762332      0
#> 686  -0.468474538      0
#> 687   0.365689249      0
#> 688  -0.595759463      0
#> 689   0.006284408      0
#> 690  -0.651673518      0
#> 691  -0.759685849      0
#> 692   0.703786197      0
#> 693  -0.250229790      0
#> 694   0.718005789      0
#> 695  -0.315056229      0
#> 696  -0.366535026      0
#> 697  -0.121721842      0
#> 698  -1.880802048      0
#> 699  -2.792711334      0
#> 700   0.198199054      0
#> 701   1.680875610      0
#> 702  -0.255728628      0
#> 703   1.282830616      0
#> 704   0.175189359      0
#> 705  -0.953277876      0
#> 706  -0.444018616      0
#> 707   0.445193774      0
#> 708   1.853190539      0
#> 709   1.458736581      0
#> 710   1.275487218      0
#> 711   1.210591707      0
#> 712  -0.473054108      0
#> 713  -0.783197163      0
#> 714   0.151278637      0
#> 715  -1.394298761      0
#> 716   0.504383044      0
#> 717  -0.799282250      0
#> 718   0.364159411      0
#> 719   0.420685435      0
#> 720  -2.138296637      0
#> 721   1.698734856      0
#> 722   0.063210832      0
#> 723  -0.051148046      0
#> 724   0.924321827      0
#> 725  -0.999423031      0
#> 726   1.563113319      0
#> 727  -0.629619466      0
#> 728   0.166973042      0
#> 729  -0.384416088      0
#> 730   1.077870826      0
#> 731  -0.073765241      0
#> 732   0.094707065      0
#> 733  -0.997443128      0
#> 734   2.299179945      0
#> 735  -0.931231939      0
#> 736   0.304877498      0
#> 737   0.213198112      0
#> 738  -0.966939696      0
#> 739   0.203712998      0
#> 740   1.566280677      0
#> 741   1.708059305      0
#> 742   0.649203326      0
#> 743  -0.998343800      0
#> 744  -0.211488704      0
#> 745  -0.268858877      0
#> 746   0.751897875      0
#> 747  -0.009273057      0
#> 748  -0.242432304      0
#> 749  -0.422446073      0
#> 750  -1.479946604      0
#> 751   0.175332268      0
#> 752   0.215259556      0
#> 753   1.171628293      0
#> 754  -0.096448725      0
#> 755  -0.511806200      0
#> 756   0.109135810      0
#> 757   0.045533089      0
#> 758  -1.606976568      0
#> 759  -0.386547338      0
#> 760  -0.894966260      0
#> 761   1.532686794      0
#> 762   0.285850307      0
#> 763   1.043322574      0
#> 764   0.084502565      0
#> 765   1.506110753      0
#> 766   1.717781228      0
#> 767  -1.201559752      0
#> 768   0.856663080      0
#> 769   0.091047557      0
#> 770   0.798051194      0
#> 771   1.942567553      0
#> 772  -1.807634647      0
#> 773   1.331710375      0
#> 774  -2.551267311      0
#> 775  -0.170566489      0
#> 776   0.733780461      0
#> 777   0.304100946      0
#> 778   0.575416796      0
#> 779  -1.356641224      0
#> 780  -0.026324905      0
#> 781   0.603514339      0
#> 782   0.451752336      0
#> 783   2.056413827      0
#> 784   1.283058597      0
#> 785   1.401335518      0
#> 786   0.182526343      0
#> 787  -1.609500888      0
#> 788   0.562890853      0
#> 789  -0.046032646      0
#> 790  -0.755858640      0
#> 791  -1.271483161      0
#> 792  -0.565799736      0
#> 793  -0.863503514      0
#> 794  -0.944853631      0
#> 795   0.370433718      0
#> 796   1.092202961      0
#> 797   1.133789548      0
#> 798  -0.922532489      0
#> 799   0.763542761      0
#> 800   0.705993387      0
#> 801  -1.555016013      0
#> 802   0.523649824      0
#> 803   0.001877669      0
#> 804   1.147839427      0
#> 805  -1.352587246      0
#> 806   0.858363662      0
#> 807  -0.360691757      0
#> 808  -1.060985382      0
#> 809  -1.414684949      0
#> 810  -0.325616643      0
#> 811   0.904399225      0
#> 812   0.781520763      0
#> 813   0.635960184      0
#> 814   0.837921347      0
#> 815   1.026943164      0
#> 816   0.904051944      0
#> 817  -0.520068974      0
#> 818  -0.434701648      0
#> 819   1.341133425      0
#> 820   0.516259704      0
#> 821   1.635561506      0
#> 822  -0.489727523      0
#> 823   0.511044896      0
#> 824   0.561365755      0
#> 825   0.949147568      0
#> 826   0.971681938      0
#> 827   1.704686567      0
#> 828   1.206085642      0
#> 829  -1.479484550      0
#> 830   0.400783121      0
#> 831  -0.771128681      0
#> 832  -1.297293916      0
#> 833   0.838967235      0
#> 834  -1.762714318      0
#> 835  -0.879232212      0
#> 836  -2.363335766      0
#> 837   2.229400608      0
#> 838   1.200864966      0
#> 839   1.258564875      0
#> 840  -0.056648329      0
#> 841   0.066459619      0
#> 842   1.589512594      0
#> 843  -1.096912860      0
#> 844   0.415773902      0
#> 845  -1.576761280      0
#> 846  -0.262687008      0
#> 847  -0.320416273      0
#> 848   0.181992980      0
#> 849   0.291598021      0
#> 850   0.274928988      0
#> 851   0.181075552      0
#> 852  -0.218924677      0
#> 853   0.564289802      0
#> 854  -0.109914801      0
#> 855  -0.632056701      0
#> 856  -0.325093765      0
#> 857  -0.626072016      0
#> 858  -0.917720489      0
#> 859   0.381406447      0
#> 860  -1.635581790      0
#> 861  -0.792121005      0
#> 862  -0.830900180      0
#> 863  -2.028234398      0
#> 864   0.472579001      0
#> 865   0.742704424      0
#> 866  -0.233598280      0
#> 867   0.116576921      0
#> 868  -0.161976076      0
#> 869  -0.596584729      0
#> 870   0.958658743      0
#> 871   0.215865512      0
#> 872   0.964992650      0
#> 873  -0.153476602      0
#> 874  -1.503231510      0
#> 875   1.276844834      0
#> 876   1.217591222      0
#> 877   1.029303749      0
#> 878  -0.639710700      0
#> 879   0.451150831      0
#> 880   0.068181187      0
#> 881   0.668494132      0
#> 882   0.930734649      0
#> 883   1.835188291      0
#> 884  -0.699889183      0
#> 885  -0.407553449      0
#> 886   1.167346157      0
#> 887   0.791716364      0
#> 888   0.866639549      0
#> 889   0.091459246      0
#> 890  -1.929938768      0
#> 891   0.755021819      0
#> 892   1.035928392      0
#> 893   1.251174377      0
#> 894   0.367673994      0
#> 895  -0.328517417      0
#> 896   0.117268747      0
#> 897  -0.355750896      0
#> 898   0.291753436      0
#> 899   0.392895610      0
#> 900   0.863242338      0
#> 901  -0.967119605      0
#> 902  -1.567897661      0
#> 903   0.913662434      0
#> 904   0.683470151      0
#> 905   0.848356792      0
#> 906   1.137429365      1
#> 907   2.149352308      1
#> 908   0.339117468      1
#> 909   0.423866440      1
#> 910   2.060910154      1
#> 911  -1.232022528      1
#> 912   1.436760154      1
#> 913   1.441375376      1
#> 914   0.283468316      1
#> 915   1.076195790      1
#> 916   1.403795018      1
#> 917   1.733778359      1
#> 918   2.295587701      1
#> 919  -0.103570077      1
#> 920   2.323204253      1
#> 921   2.044279699      1
#> 922  -0.645693417      1
#> 923   2.013812615      1
#> 924   1.365610517      1
#> 925   0.620919235      1
#> 926   1.141665938      1
#> 927  -0.646197383      1
#> 928   2.399438123      1
#> 929   1.166636066      1
#> 930   1.077996116      1
#> 931   0.716324167      1
#> 932   1.043394847      1
#> 933   0.355934626      1
#> 934  -0.217533906      1
#> 935   1.092855576      1
#> 936  -0.129945324      1
#> 937   1.667224429      1
#> 938  -1.406145774      1
#> 939   0.603871710      1
#> 940   1.233963314      1
#> 941   1.164677037      1
#> 942   1.497261634      1
#> 943   1.044239379      1
#> 944   0.625624119      1
#> 945  -1.062902176      1
#> 946   0.978911894      1
#> 947   0.987395434      1
#> 948   1.136808641      1
#> 949   2.664717448      1
#> 950   1.991322885      1
#> 951  -1.880600081      1
#> 952   0.760537099      1
#> 953   0.632994612      1
#> 954   1.593890362      1
#> 955   1.349317125      1
#> 956   0.571150145      1
#> 957   2.029262638      1
#> 958   0.548711135      1
#> 959  -0.067230523      1
#> 960  -0.673937100      1
#> 961   0.189206725      1
#> 962   2.729886239      1
#> 963   1.079636916      1
#> 964   0.109551675      1
#> 965   1.115743460      1
#> 966   0.178079812      1
#> 967   2.054789699      1
#> 968  -0.201638387      1
#> 969   1.877091390      1
#> 970   1.510156183      1
#> 971   1.938884098      1
#> 972   2.404699533      1
#> 973   0.907323343      1
#> 974   1.500629872      1
#> 975  -0.045906531      1
#> 976   1.063068412      1
#> 977   0.499292705      1
#> 978   0.382229156      1
#> 979   1.373430994      1
#> 980   0.435069042      1
#> 981   1.140615500      1
#> 982  -0.487503094      1
#> 983  -1.049113110      1
#> 984   0.749453785      1
#> 985   2.394815096      1
#> 986   0.864124952      1
#> 987   1.808931315      1
#> 988   2.416522468      1
#> 989   1.130624857      1
#> 990   1.338079506      1
#> 991   1.387247933      1
#> 992   0.928208935      1
#> 993   0.961365906      1
#> 994  -1.115988346      1
#> 995   0.884346681      1
#> 996   0.922128340      1
#> 997   2.096462784      1
#> 998   0.028134777      1
#> 999  -0.228779778      1
#> 1000 -0.401933027      1