Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.245369701      0
#> 2    -1.188716705      0
#> 3     0.205582473      0
#> 4    -1.280689605      0
#> 5     0.061235743      0
#> 6     0.055791008      0
#> 7    -2.747379916      0
#> 8    -0.449094562      0
#> 9    -0.855000984      0
#> 10   -0.810499884      0
#> 11   -0.533294363      0
#> 12    0.082764660      0
#> 13   -0.004915498      0
#> 14    1.024828553      0
#> 15   -0.732563654      0
#> 16    0.789460396      0
#> 17   -1.455042607      0
#> 18   -0.164559220      0
#> 19    1.123539183      0
#> 20   -0.158230977      0
#> 21   -0.452009694      0
#> 22   -1.307923197      0
#> 23   -0.602132365      0
#> 24    0.056133626      0
#> 25   -0.300842308      0
#> 26    0.894745530      0
#> 27   -0.926977227      0
#> 28   -0.099236319      0
#> 29   -0.533037665      0
#> 30   -0.201267415      0
#> 31    0.982909825      0
#> 32    0.423973479      0
#> 33    1.211467971      0
#> 34    0.018400970      0
#> 35   -0.691280997      0
#> 36    0.503006771      0
#> 37    0.181084185      0
#> 38   -0.204487929      0
#> 39   -1.069466201      0
#> 40   -0.028661620      0
#> 41    0.739957407      0
#> 42   -0.128175065      0
#> 43   -0.706699866      0
#> 44    1.465782581      0
#> 45   -1.540453086      0
#> 46   -0.302042539      0
#> 47   -0.198654720      0
#> 48   -0.265400070      0
#> 49   -0.511063435      0
#> 50    1.231023732      0
#> 51   -1.127340348      0
#> 52   -0.608313534      0
#> 53   -0.136420875      0
#> 54   -0.670541757      0
#> 55   -0.174859531      0
#> 56    0.209909587      0
#> 57   -0.927569043      0
#> 58   -0.401964330      0
#> 59    0.446559862      0
#> 60   -0.054251294      0
#> 61   -0.544292476      0
#> 62    0.010912042      0
#> 63    1.592793043      0
#> 64   -0.342327731      0
#> 65    1.354430463      0
#> 66   -0.968333578      0
#> 67    0.015322266      0
#> 68    0.996436518      0
#> 69    2.451614452      0
#> 70   -0.307200661      0
#> 71    0.033775044      0
#> 72    1.739139014      0
#> 73    0.028382095      0
#> 74    1.532416793      0
#> 75   -1.101416569      0
#> 76   -0.379849014      0
#> 77   -0.047457741      0
#> 78   -0.597381138      0
#> 79    2.970247863      0
#> 80    0.559889614      0
#> 81   -0.079408511      0
#> 82    0.555024007      0
#> 83   -0.138509736      0
#> 84    0.325339252      0
#> 85    2.219150427      0
#> 86   -0.352264365      0
#> 87   -0.005719538      0
#> 88   -0.887144171      0
#> 89    0.112932916      0
#> 90    1.021164260      0
#> 91    0.392599375      0
#> 92   -0.851705431      0
#> 93   -2.350470764      0
#> 94    0.418996647      0
#> 95   -0.646936588      0
#> 96   -0.172081597      0
#> 97   -0.497545682      0
#> 98    0.327900502      0
#> 99   -0.013768069      0
#> 100   1.949705420      0
#> 101  -1.028785885      0
#> 102   0.046030794      0
#> 103  -0.310752495      0
#> 104  -0.660190992      0
#> 105   0.977105760      0
#> 106   0.161657399      0
#> 107  -0.803591390      0
#> 108   0.738326503      0
#> 109  -1.669913963      0
#> 110  -0.443883100      0
#> 111   0.470008090      0
#> 112   0.845899143      0
#> 113   0.942467989      0
#> 114  -1.141361347      0
#> 115   0.647946152      0
#> 116  -0.059436509      0
#> 117  -0.921186732      0
#> 118   0.542689239      0
#> 119  -0.428464107      0
#> 120  -0.684140495      0
#> 121  -2.310767191      0
#> 122   0.563497899      0
#> 123   0.830765333      0
#> 124  -0.765694527      0
#> 125  -1.214828073      0
#> 126   0.994035142      0
#> 127   0.738045855      0
#> 128   0.808135429      0
#> 129   0.830310712      0
#> 130  -0.515699448      0
#> 131  -0.167863369      0
#> 132   0.757609415      0
#> 133  -1.687873728      0
#> 134  -0.368410729      0
#> 135   0.826159253      0
#> 136  -2.180185280      0
#> 137  -0.531258470      0
#> 138   0.228280336      0
#> 139  -2.584607811      0
#> 140  -0.497623815      0
#> 141   0.785432761      0
#> 142  -0.008830416      0
#> 143  -0.657847270      0
#> 144  -0.344546255      0
#> 145  -0.234912832      0
#> 146   2.808043711      0
#> 147  -0.153626147      0
#> 148  -1.168877747      0
#> 149  -0.372927672      0
#> 150  -0.664572315      0
#> 151   0.953450441      0
#> 152  -0.572508540      0
#> 153  -0.965159785      0
#> 154   1.588465795      0
#> 155  -0.778305352      0
#> 156   2.319077035      0
#> 157  -0.678717530      0
#> 158   0.082778249      0
#> 159   0.148317363      0
#> 160  -1.338108056      0
#> 161   0.537939642      0
#> 162  -0.326746707      0
#> 163   2.035486482      0
#> 164   0.552142702      0
#> 165  -1.146750595      0
#> 166   2.109004324      0
#> 167  -1.676295317      0
#> 168  -1.003978337      0
#> 169  -1.733800924      0
#> 170  -1.203750935      0
#> 171   0.946754931      0
#> 172  -1.353977947      0
#> 173  -0.407432090      0
#> 174  -0.713488713      0
#> 175   0.145365525      0
#> 176   0.854449984      0
#> 177  -0.596276368      0
#> 178   1.786906124      0
#> 179  -0.220725603      0
#> 180  -0.267272608      0
#> 181  -0.566757773      0
#> 182  -0.432121316      0
#> 183   0.573292876      0
#> 184   0.560252144      0
#> 185   0.580544233      0
#> 186   0.438099955      0
#> 187   0.707684611      0
#> 188   1.059580147      0
#> 189   1.067398569      0
#> 190  -0.493171329      0
#> 191   0.381323928      0
#> 192   0.212419378      0
#> 193  -1.764702238      0
#> 194   0.375433144      0
#> 195   0.495973256      0
#> 196   0.665803775      0
#> 197  -1.260414987      0
#> 198  -2.670058157      0
#> 199  -1.045911245      0
#> 200  -0.602926383      0
#> 201   0.808304300      0
#> 202   0.379921987      0
#> 203   0.146138626      0
#> 204   0.044724083      0
#> 205  -1.470592746      0
#> 206   0.736831039      0
#> 207   1.582595764      0
#> 208   1.447823759      0
#> 209  -3.194347014      0
#> 210   1.078899147      0
#> 211  -0.103228393      0
#> 212  -0.554177163      0
#> 213   0.303957632      0
#> 214   0.274482764      0
#> 215  -1.176166559      0
#> 216   0.107600200      0
#> 217  -1.184683384      0
#> 218   0.339372333      0
#> 219  -1.292991861      0
#> 220   0.316315031      0
#> 221  -0.089773109      0
#> 222  -1.229201198      0
#> 223   1.428312165      0
#> 224  -0.894180680      0
#> 225  -0.256177956      0
#> 226   0.242157214      0
#> 227   2.579493430      0
#> 228  -0.571175523      0
#> 229   2.214910293      0
#> 230   0.101508323      0
#> 231   0.081216910      0
#> 232  -1.142925897      0
#> 233   0.138722020      0
#> 234   1.486101178      0
#> 235  -0.605904039      0
#> 236   1.104021866      0
#> 237  -1.251130962      0
#> 238   0.963239627      0
#> 239  -0.035722363      0
#> 240   0.094322896      0
#> 241  -0.154520283      0
#> 242   0.965817289      0
#> 243  -2.265089522      0
#> 244   0.747213695      0
#> 245   1.056459047      0
#> 246  -0.236757784      0
#> 247   0.360586167      0
#> 248   0.747538668      0
#> 249  -2.078956458      0
#> 250  -1.780775459      0
#> 251   1.628393594      0
#> 252  -1.244833273      0
#> 253  -1.511006214      0
#> 254  -1.632819644      0
#> 255  -1.587108383      0
#> 256   0.900797082      0
#> 257   0.883641975      0
#> 258   0.155644907      0
#> 259   0.792593697      0
#> 260   0.272334275      0
#> 261  -0.852251236      0
#> 262  -0.782861112      0
#> 263   1.700982343      0
#> 264   0.998553962      0
#> 265   1.364654381      0
#> 266   0.223334440      0
#> 267  -0.290539237      0
#> 268   0.376723910      0
#> 269   0.163925877      0
#> 270  -0.758159506      0
#> 271  -0.622761503      0
#> 272   0.755285829      0
#> 273   0.115488360      0
#> 274  -1.375511733      0
#> 275   0.644444850      0
#> 276  -0.534551564      0
#> 277  -0.192196440      0
#> 278  -0.433200352      0
#> 279   0.171700173      0
#> 280   0.127677770      0
#> 281   0.022805739      0
#> 282   2.534728082      0
#> 283  -0.848732335      0
#> 284   0.964926005      0
#> 285  -1.754554082      0
#> 286  -0.158838522      0
#> 287   1.631041469      0
#> 288  -0.734055321      0
#> 289  -0.560994030      0
#> 290  -0.352040998      0
#> 291   0.490486057      0
#> 292  -0.934606832      0
#> 293  -1.278270879      0
#> 294  -0.532053152      0
#> 295  -1.589069124      0
#> 296   0.940728999      0
#> 297   1.487154809      0
#> 298   0.594363500      0
#> 299  -1.409562702      0
#> 300   0.755538736      0
#> 301   0.232827934      0
#> 302   1.104568024      0
#> 303   0.386583045      0
#> 304  -0.610686748      0
#> 305   0.189468883      0
#> 306   0.660827252      0
#> 307   1.446901337      0
#> 308  -1.326442514      0
#> 309  -0.880954492      0
#> 310  -2.556380505      0
#> 311   0.109561442      0
#> 312  -1.023904413      0
#> 313  -1.158574593      0
#> 314   0.772576211      0
#> 315  -0.184630153      0
#> 316  -1.129975847      0
#> 317   0.730442088      0
#> 318   0.717792855      0
#> 319  -1.282931884      0
#> 320   0.600287315      0
#> 321   1.027219529      0
#> 322  -1.868262833      0
#> 323  -1.182154192      0
#> 324   0.855206465      0
#> 325  -0.382236334      0
#> 326   0.019178660      0
#> 327   1.537717245      0
#> 328  -1.870607629      0
#> 329   1.018393656      0
#> 330  -0.704956853      0
#> 331  -1.075212414      0
#> 332  -0.051850093      0
#> 333   3.682219931      0
#> 334  -1.221494608      0
#> 335  -0.801107693      0
#> 336  -1.090129426      0
#> 337  -0.200125422      0
#> 338   0.863149116      0
#> 339   0.050646710      0
#> 340  -0.358141688      0
#> 341  -0.200532235      0
#> 342  -0.487911369      0
#> 343  -0.443334913      0
#> 344   0.728640925      0
#> 345  -0.617365681      0
#> 346  -3.003167506      0
#> 347  -0.765648196      0
#> 348  -0.778932072      0
#> 349   1.433246178      0
#> 350  -0.110915841      0
#> 351   0.039953309      0
#> 352   1.358929530      0
#> 353  -0.559867789      0
#> 354  -0.210252856      0
#> 355   0.831240242      0
#> 356  -0.073376568      0
#> 357   0.159828041      0
#> 358   0.897561655      0
#> 359   0.607485656      0
#> 360  -1.660925553      0
#> 361  -0.811968888      0
#> 362  -1.288349458      0
#> 363  -1.027353529      0
#> 364   0.217957881      0
#> 365   1.775414905      0
#> 366   1.183270153      0
#> 367  -0.795552318      0
#> 368  -1.356997963      0
#> 369  -0.089971035      0
#> 370  -1.437490711      0
#> 371  -0.574524071      0
#> 372   0.831834400      0
#> 373  -0.043386420      0
#> 374   0.466113069      0
#> 375  -0.898943337      0
#> 376   1.091609396      0
#> 377  -1.179409195      0
#> 378  -1.170314863      0
#> 379  -0.595063308      0
#> 380   0.103316178      0
#> 381  -0.024842537      0
#> 382  -0.083264527      0
#> 383   0.303669547      0
#> 384  -0.179434313      0
#> 385   0.222593684      0
#> 386  -1.680117317      0
#> 387   1.278828804      0
#> 388  -0.486476879      0
#> 389   0.729903964      0
#> 390  -1.159472462      0
#> 391   0.426422260      0
#> 392   1.385097227      0
#> 393   0.029715888      0
#> 394  -0.028989634      0
#> 395   0.432186203      0
#> 396  -0.336660582      0
#> 397   1.199860179      0
#> 398   0.045638147      0
#> 399   2.060774618      0
#> 400   0.188913290      0
#> 401  -0.513845955      0
#> 402  -0.679462882      0
#> 403   1.503893838      0
#> 404   2.070908335      0
#> 405   1.334813656      0
#> 406  -0.521735233      0
#> 407   0.428692798      0
#> 408  -0.135852235      0
#> 409   2.348470231      0
#> 410  -0.209121888      0
#> 411  -2.267948231      0
#> 412   0.533961263      0
#> 413   0.339686651      0
#> 414   0.087483012      0
#> 415  -0.729274684      0
#> 416   0.923635114      0
#> 417   0.941647039      0
#> 418   0.328603634      0
#> 419  -0.795579378      0
#> 420  -1.878019005      0
#> 421   1.599662850      0
#> 422  -0.423218267      0
#> 423  -0.167700181      0
#> 424  -0.333860576      0
#> 425  -0.538487994      0
#> 426  -0.098490165      0
#> 427  -0.202063584      0
#> 428   0.662337465      0
#> 429  -0.292189193      0
#> 430   0.446703328      0
#> 431   1.164168215      0
#> 432   0.213765181      0
#> 433   1.069176525      0
#> 434   2.572031508      0
#> 435  -0.951753984      0
#> 436  -0.172462287      0
#> 437   1.189069313      0
#> 438  -0.564511076      0
#> 439  -0.393387541      0
#> 440   0.345589784      0
#> 441  -1.477401108      0
#> 442   0.576997769      0
#> 443  -0.320547420      0
#> 444  -0.885760428      0
#> 445  -0.225768940      0
#> 446   1.391745982      0
#> 447  -0.844281332      0
#> 448   0.700562920      0
#> 449  -2.134506806      0
#> 450  -0.994877155      0
#> 451   0.264711467      0
#> 452  -0.275509365      0
#> 453  -0.188620800      0
#> 454  -0.730145673      0
#> 455  -1.146707637      0
#> 456  -0.616780655      0
#> 457  -0.405252784      0
#> 458   0.237925816      0
#> 459  -1.209227217      0
#> 460  -0.950577055      0
#> 461   0.583025038      0
#> 462  -0.373041012      0
#> 463  -1.016802378      0
#> 464  -1.333312009      0
#> 465  -1.009872263      0
#> 466  -1.741663970      0
#> 467  -1.915523557      0
#> 468  -0.221915958      0
#> 469   0.852573914      0
#> 470  -0.607907544      0
#> 471   0.720048679      0
#> 472  -1.656984166      0
#> 473   0.180471862      0
#> 474   0.442352189      0
#> 475   2.245568717      0
#> 476   0.367626341      0
#> 477   1.172101459      0
#> 478  -0.469449665      0
#> 479   1.205312335      0
#> 480   0.318592982      0
#> 481  -0.028137826      0
#> 482   0.926738987      0
#> 483  -0.216910887      0
#> 484   0.591108044      0
#> 485   0.416391078      0
#> 486   0.438156018      0
#> 487   0.659718725      0
#> 488  -1.401313911      0
#> 489   0.637999604      0
#> 490  -1.513338614      0
#> 491   0.404229149      0
#> 492  -0.132432155      0
#> 493   0.402022921      0
#> 494  -0.712848185      0
#> 495   0.252216897      0
#> 496   0.844915655      0
#> 497   0.470966312      0
#> 498  -0.306977128      0
#> 499   0.070229962      0
#> 500   0.761675574      0
#> 501  -2.100165848      0
#> 502  -0.038635314      0
#> 503   0.189370836      0
#> 504  -0.246503347      0
#> 505   0.461853853      0
#> 506   0.183430473      0
#> 507   2.363035132      0
#> 508   2.182724105      0
#> 509  -2.683708929      0
#> 510   2.289366140      0
#> 511  -0.824203365      0
#> 512   0.656895353      0
#> 513  -1.670043321      0
#> 514   0.666658488      0
#> 515   0.797048733      0
#> 516   1.621013306      0
#> 517  -0.504461918      0
#> 518  -2.556400225      0
#> 519  -0.255667673      0
#> 520   1.084459657      0
#> 521   0.366074802      0
#> 522  -0.803031280      0
#> 523  -1.399312987      0
#> 524  -1.288929495      0
#> 525  -1.218161440      0
#> 526  -0.487168767      0
#> 527  -1.021529606      0
#> 528  -0.297938083      0
#> 529   0.912364202      0
#> 530   1.467597310      0
#> 531  -1.087352585      0
#> 532  -0.135106437      0
#> 533  -0.774979640      0
#> 534   1.397736639      0
#> 535  -1.617325708      0
#> 536  -0.508211753      0
#> 537  -0.220494124      0
#> 538  -0.795436480      0
#> 539   0.432777664      0
#> 540  -1.218873574      0
#> 541  -0.254389666      0
#> 542   0.334458283      0
#> 543  -0.489744426      0
#> 544   0.329741216      0
#> 545   1.173846921      0
#> 546   1.320313175      0
#> 547   1.418090971      0
#> 548   0.908157760      0
#> 549   0.163316105      0
#> 550   0.305399133      0
#> 551   0.897190499      0
#> 552  -1.857809619      0
#> 553  -0.847792725      0
#> 554  -2.499568636      0
#> 555  -0.797277855      0
#> 556   0.131799083      0
#> 557   0.084448535      0
#> 558   0.055860605      0
#> 559   0.026543409      0
#> 560  -1.486744137      0
#> 561   1.336479392      0
#> 562   1.011593939      0
#> 563   0.598055488      0
#> 564   0.931110341      0
#> 565   0.957535986      0
#> 566   0.615493122      0
#> 567   1.858612923      0
#> 568  -0.400795339      0
#> 569   0.481975482      0
#> 570  -0.377576004      0
#> 571  -1.282289490      0
#> 572  -0.421191281      0
#> 573  -0.633912231      0
#> 574   0.315582840      0
#> 575   0.279962411      0
#> 576  -1.553927198      0
#> 577  -0.053653260      0
#> 578   0.749542495      0
#> 579   0.768854976      0
#> 580   1.660405410      0
#> 581   1.071429561      0
#> 582   1.129749655      0
#> 583  -0.393607908      0
#> 584  -0.918251767      0
#> 585   0.286458564      0
#> 586  -3.434783196      0
#> 587  -1.531852926      0
#> 588   0.040086493      0
#> 589  -0.464853807      0
#> 590   0.229387272      0
#> 591   0.047731416      0
#> 592  -1.424318312      0
#> 593   0.324357260      0
#> 594  -1.262702577      0
#> 595   1.363917354      0
#> 596   0.771090630      0
#> 597  -0.545350955      0
#> 598  -1.895416260      0
#> 599   0.623631744      0
#> 600  -1.363359243      0
#> 601  -1.241156920      0
#> 602   0.193980760      0
#> 603  -0.058492912      0
#> 604   1.502812988      0
#> 605  -0.904651891      0
#> 606   1.046499374      0
#> 607  -0.098130354      0
#> 608   0.181695924      0
#> 609  -0.131832235      0
#> 610   1.610269777      0
#> 611  -0.622379866      0
#> 612  -0.542872838      0
#> 613  -0.579385982      0
#> 614   0.883980320      0
#> 615   0.043094223      0
#> 616  -0.593641614      0
#> 617   0.104690508      0
#> 618   0.325217733      0
#> 619  -1.066961073      0
#> 620   0.495732931      0
#> 621  -0.194815332      0
#> 622   0.021328143      0
#> 623  -1.671180933      0
#> 624  -2.202234218      0
#> 625   1.325886735      0
#> 626  -0.376101181      0
#> 627   0.351149137      0
#> 628   0.412817788      0
#> 629   1.025966409      0
#> 630  -0.450585830      0
#> 631  -0.433764928      0
#> 632   0.727282616      0
#> 633  -1.276621415      0
#> 634   0.300545350      0
#> 635   0.070220554      0
#> 636  -0.899606139      0
#> 637   0.149202150      0
#> 638   0.228862378      0
#> 639  -0.431620348      0
#> 640   0.738995635      0
#> 641  -0.325089762      0
#> 642   1.367029253      0
#> 643   0.726429790      0
#> 644   0.858232371      0
#> 645   0.558443776      0
#> 646  -1.514622940      0
#> 647   2.471334514      0
#> 648   0.990177046      0
#> 649  -0.835340168      0
#> 650   0.093652518      0
#> 651   0.762318713      0
#> 652   0.184567653      0
#> 653  -0.249598801      0
#> 654   1.250863613      0
#> 655   0.447163134      0
#> 656   2.023069409      0
#> 657  -0.740233271      0
#> 658   0.763557454      0
#> 659  -1.659832485      0
#> 660  -0.176772401      0
#> 661  -1.091098551      0
#> 662  -2.029685454      0
#> 663   1.448989979      0
#> 664  -0.170778089      0
#> 665   0.799677060      0
#> 666  -0.029305826      0
#> 667   0.809574695      0
#> 668   1.037765101      0
#> 669   0.393452648      0
#> 670   0.207000662      0
#> 671   0.660303077      0
#> 672  -1.777050360      0
#> 673   1.230967943      0
#> 674  -0.697178970      0
#> 675   2.446953650      0
#> 676  -1.607442059      0
#> 677  -0.846213085      0
#> 678   0.433602878      0
#> 679  -0.145067786      0
#> 680   0.678842013      0
#> 681  -1.247955938      0
#> 682  -0.557954255      0
#> 683  -0.583996538      0
#> 684  -0.191514100      0
#> 685  -0.624036128      0
#> 686  -0.296499937      0
#> 687   0.417254676      0
#> 688   1.080001843      0
#> 689  -1.491744404      0
#> 690   0.003306829      0
#> 691  -0.201547771      0
#> 692   2.409491275      0
#> 693   1.166141592      0
#> 694  -0.508902638      0
#> 695  -0.292772092      0
#> 696   0.163779059      0
#> 697   0.579349254      0
#> 698   0.985311474      0
#> 699   1.149986662      0
#> 700  -0.941617769      0
#> 701   0.365942435      0
#> 702  -0.622001714      0
#> 703   0.796228653      0
#> 704   0.004582561      0
#> 705   1.236687189      0
#> 706   0.974965107      0
#> 707  -0.269343523      0
#> 708  -1.403482944      0
#> 709   2.139326963      0
#> 710   0.833413757      0
#> 711  -0.580176611      0
#> 712  -0.407129098      0
#> 713   0.898781733      0
#> 714  -1.994275176      0
#> 715   0.687030927      0
#> 716  -1.050953123      0
#> 717   0.494272411      0
#> 718  -1.117838931      0
#> 719   0.569750798      0
#> 720   0.788886177      0
#> 721  -1.024120305      0
#> 722   0.391015470      0
#> 723  -0.674923958      0
#> 724  -1.577042913      0
#> 725   0.609261414      0
#> 726   1.067283106      0
#> 727  -0.380107190      0
#> 728  -0.901609634      0
#> 729  -1.914520729      0
#> 730  -1.658613561      0
#> 731  -1.640720610      0
#> 732  -0.969781919      0
#> 733  -0.283656790      0
#> 734  -0.132100121      0
#> 735  -1.269671022      0
#> 736  -0.818386009      0
#> 737  -0.163551234      0
#> 738   0.082083058      0
#> 739   0.076476421      0
#> 740   0.473329551      0
#> 741  -0.722678223      0
#> 742   0.298168790      0
#> 743   0.482573567      0
#> 744  -0.360594108      0
#> 745   0.740953747      0
#> 746   0.808077866      0
#> 747  -1.290711632      0
#> 748   0.503357273      0
#> 749   0.197278283      0
#> 750  -0.846228581      0
#> 751   0.944134684      0
#> 752  -0.200004792      0
#> 753   0.156712734      0
#> 754   0.059460852      0
#> 755   1.709211938      0
#> 756  -0.780413976      0
#> 757  -3.326388185      0
#> 758   0.993166109      0
#> 759  -0.082657565      0
#> 760  -0.446085369      0
#> 761  -0.113752001      0
#> 762   0.545592710      0
#> 763  -0.467566217      0
#> 764  -0.184366631      0
#> 765  -1.518098340      0
#> 766  -0.628773643      0
#> 767   0.867476114      0
#> 768   0.215255807      0
#> 769   1.885187375      0
#> 770   0.198704943      0
#> 771  -0.807219727      0
#> 772  -1.266167622      0
#> 773   1.174533964      0
#> 774  -0.472754409      0
#> 775   0.326500891      0
#> 776  -0.356864000      0
#> 777   0.036248223      0
#> 778  -0.105041410      0
#> 779   0.112209470      0
#> 780   1.695849670      0
#> 781  -0.771324665      0
#> 782  -1.555784823      0
#> 783   1.001724107      0
#> 784   1.080332943      0
#> 785  -1.185469350      0
#> 786  -0.715022190      0
#> 787  -0.190588753      0
#> 788  -0.666770683      0
#> 789  -0.718006111      0
#> 790   1.048423654      0
#> 791  -0.090193413      0
#> 792   0.473669351      0
#> 793  -0.550648942      0
#> 794   0.108532625      0
#> 795   0.226929517      0
#> 796  -0.673670245      0
#> 797  -0.015808929      0
#> 798   0.045788082      0
#> 799  -0.717396949      0
#> 800   0.913940120      0
#> 801  -2.624281076      0
#> 802   0.191567393      0
#> 803   0.893208639      0
#> 804   2.087130982      0
#> 805  -0.344288171      0
#> 806  -1.205840745      0
#> 807  -1.286718797      0
#> 808   0.364693955      0
#> 809  -0.500079842      0
#> 810   0.917060030      0
#> 811  -3.145250167      0
#> 812  -1.603199515      0
#> 813  -0.627977337      0
#> 814  -0.282552434      0
#> 815   0.857573514      0
#> 816  -0.962853724      0
#> 817   1.447116231      0
#> 818  -1.350385893      0
#> 819   1.208776531      0
#> 820  -1.326437098      0
#> 821  -0.945723152      0
#> 822  -1.185429666      0
#> 823  -1.561270320      0
#> 824  -2.864972711      0
#> 825  -1.300096487      0
#> 826  -0.642793763      0
#> 827   1.357608552      0
#> 828   1.830075977      0
#> 829  -0.409557649      0
#> 830  -0.666764005      0
#> 831   0.812320386      0
#> 832   0.749457501      0
#> 833  -0.314739521      0
#> 834  -0.830412947      0
#> 835   1.685814619      0
#> 836   1.283152761      0
#> 837  -1.283947327      0
#> 838  -0.154022251      0
#> 839  -0.905376464      0
#> 840  -0.873589920      0
#> 841  -0.937888097      0
#> 842   0.556616375      0
#> 843  -1.287859211      0
#> 844  -0.108815667      0
#> 845   1.169578655      0
#> 846  -1.633553091      0
#> 847  -1.232380137      0
#> 848  -1.171024224      0
#> 849  -0.320367613      0
#> 850   0.670338588      0
#> 851  -0.760162262      0
#> 852   1.834294153      0
#> 853  -0.530917034      0
#> 854   1.313513898      0
#> 855   0.650160307      0
#> 856  -0.833368102      0
#> 857  -0.146947904      0
#> 858   1.298611746      0
#> 859  -0.094941312      0
#> 860   0.881170801      0
#> 861   0.088062624      0
#> 862   0.250301192      0
#> 863   0.643105861      0
#> 864   0.220945939      0
#> 865  -0.657068723      0
#> 866   0.500377507      0
#> 867   1.831631052      0
#> 868   0.157249114      0
#> 869  -0.403201779      0
#> 870   2.055631354      0
#> 871   0.043412070      0
#> 872  -1.255947336      0
#> 873  -0.601499438      0
#> 874  -1.427687144      0
#> 875   1.838028985      0
#> 876  -0.415532198      0
#> 877  -0.190259534      0
#> 878  -2.178197307      0
#> 879  -0.834759580      0
#> 880   1.006187632      0
#> 881  -0.177133917      0
#> 882   0.250988672      0
#> 883  -0.778000575      0
#> 884  -0.984570320      0
#> 885   1.841548172      0
#> 886   0.779315474      0
#> 887   0.412975582      0
#> 888  -0.082052697      0
#> 889   0.417609111      0
#> 890  -1.602583252      0
#> 891  -0.151956201      0
#> 892   0.678002394      0
#> 893  -2.178029360      0
#> 894   0.830484602      0
#> 895   0.136394073      0
#> 896   0.887077647      0
#> 897  -0.865670947      0
#> 898  -0.631294748      0
#> 899   1.111368461      0
#> 900   0.592173492      1
#> 901   0.613247924      1
#> 902   0.318800412      1
#> 903  -0.168240740      1
#> 904   1.386980491      1
#> 905   0.622420205      1
#> 906   0.899657862      1
#> 907   0.979846066      1
#> 908   0.300000061      1
#> 909   1.147438201      1
#> 910   1.647194141      1
#> 911   1.972132844      1
#> 912   2.433960526      1
#> 913   1.777716772      1
#> 914  -0.666106368      1
#> 915   1.378878700      1
#> 916   0.946933851      1
#> 917   0.185344116      1
#> 918  -0.564265585      1
#> 919   0.732013816      1
#> 920  -0.176323520      1
#> 921   0.307019312      1
#> 922   0.934962572      1
#> 923  -1.028407784      1
#> 924   0.160139505      1
#> 925  -0.969650819      1
#> 926   1.994554819      1
#> 927   0.996964952      1
#> 928   0.074705708      1
#> 929   1.874916819      1
#> 930   0.397080590      1
#> 931   0.015472666      1
#> 932  -1.590611620      1
#> 933   0.308078023      1
#> 934   1.340905109      1
#> 935   0.654975326      1
#> 936  -0.037933875      1
#> 937   0.307225914      1
#> 938   1.177345360      1
#> 939   1.532750940      1
#> 940   1.443667442      1
#> 941   2.321364698      1
#> 942  -0.200507381      1
#> 943   0.063746183      1
#> 944   0.656364301      1
#> 945   1.099680543      1
#> 946   0.323823212      1
#> 947   0.332950054      1
#> 948  -0.805408912      1
#> 949   1.117926068      1
#> 950  -0.533909103      1
#> 951   1.324133504      1
#> 952   0.256597594      1
#> 953  -0.618854772      1
#> 954  -0.562344340      1
#> 955  -0.141874775      1
#> 956  -0.500349808      1
#> 957   0.876839255      1
#> 958  -0.129202011      1
#> 959   0.775737177      1
#> 960   1.863824083      1
#> 961   0.471378414      1
#> 962   1.194340618      1
#> 963   0.999546310      1
#> 964   2.363856825      1
#> 965  -0.666778591      1
#> 966   1.479035153      1
#> 967   1.581166083      1
#> 968  -0.628986058      1
#> 969   1.244758487      1
#> 970  -0.714041859      1
#> 971   1.074996552      1
#> 972  -0.464533157      1
#> 973   1.251912179      1
#> 974   0.823793148      1
#> 975   1.461538339      1
#> 976   1.624231327      1
#> 977   0.328679217      1
#> 978   0.661115751      1
#> 979   0.330656371      1
#> 980   0.429313835      1
#> 981  -0.139464504      1
#> 982   3.757862342      1
#> 983   0.595683601      1
#> 984   0.393358394      1
#> 985   1.375252416      1
#> 986  -0.630358963      1
#> 987   2.988605175      1
#> 988   1.691048107      1
#> 989  -0.471479784      1
#> 990   2.447011306      1
#> 991   0.784673071      1
#> 992   0.931691897      1
#> 993   1.225425869      1
#> 994   2.201133735      1
#> 995   0.127570305      1
#> 996   0.690241060      1
#> 997  -1.486037750      1
#> 998   0.936925436      1
#> 999   0.303902162      1
#> 1000  1.622405501      1