Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     0.8849303552      0
#> 2     0.6550114810      0
#> 3     0.3541219647      0
#> 4     0.7885653005      0
#> 5     1.2735105646      0
#> 6    -0.4492104191      0
#> 7     0.6332868208      0
#> 8    -2.0989237303      0
#> 9    -0.5822386437      0
#> 10    1.3499511659      0
#> 11   -0.8368158865      0
#> 12    1.7561011911      0
#> 13    0.4061071298      0
#> 14   -0.6231076375      0
#> 15   -0.0966355000      0
#> 16   -1.8963638297      0
#> 17    0.5114258224      0
#> 18    0.4652568791      0
#> 19   -1.6429254064      0
#> 20   -1.3994438776      0
#> 21    0.2604586986      0
#> 22   -0.6779039624      0
#> 23   -0.5026385869      0
#> 24   -0.4956419219      0
#> 25    0.6431155656      0
#> 26    0.4058754619      0
#> 27    0.7720454571      0
#> 28    1.7201640052      0
#> 29   -1.3868760236      0
#> 30   -0.4591755158      0
#> 31   -0.0117384495      0
#> 32    0.7447026728      0
#> 33   -0.9037807280      0
#> 34    1.8775158804      0
#> 35   -0.2120474610      0
#> 36   -2.8373444064      0
#> 37   -0.8159046178      0
#> 38   -0.4679693074      0
#> 39    0.9908377171      0
#> 40    0.6811927034      0
#> 41    0.6589078051      0
#> 42   -2.4713877592      0
#> 43   -1.0468667092      0
#> 44    0.4696741568      0
#> 45    0.9066925025      0
#> 46   -0.0326646712      0
#> 47    0.2781591674      0
#> 48   -0.3432785539      0
#> 49    0.8652983799      0
#> 50   -0.0980044106      0
#> 51    0.4311697917      0
#> 52   -0.3057303220      0
#> 53    0.6304474231      0
#> 54   -0.3462646658      0
#> 55   -0.2758470869      0
#> 56   -0.0396279036      0
#> 57    0.0499866863      0
#> 58   -0.0960441984      0
#> 59   -1.7848398768      0
#> 60    0.4076491400      0
#> 61    0.4458187712      0
#> 62   -0.0992915909      0
#> 63   -0.9946088333      0
#> 64    0.8557528038      0
#> 65    0.1128521477      0
#> 66   -0.2715889938      0
#> 67    0.4918476384      0
#> 68    1.5241539228      0
#> 69    0.9842629901      0
#> 70    1.3204521366      0
#> 71   -0.5875536009      0
#> 72    1.8813818288      0
#> 73    0.1705295658      0
#> 74   -1.2934902948      0
#> 75   -0.6103427292      0
#> 76    0.3081275021      0
#> 77   -0.5550678098      0
#> 78    0.4098150794      0
#> 79    0.7645296914      0
#> 80    0.5455181484      0
#> 81   -1.5258453423      0
#> 82    0.4124377795      0
#> 83    1.1382704729      0
#> 84   -0.6841041302      0
#> 85    0.8661599746      0
#> 86   -0.5399566531      0
#> 87   -0.5416361601      0
#> 88   -0.1639827591      0
#> 89   -0.3988346414      0
#> 90   -1.6659385525      0
#> 91    2.3426657151      0
#> 92    1.3955444276      0
#> 93    0.6093879159      0
#> 94   -0.2341899903      0
#> 95    1.6651555343      0
#> 96    0.7099436908      0
#> 97   -1.9842624884      0
#> 98    0.3173896770      0
#> 99   -2.6100034374      0
#> 100  -1.4976350547      0
#> 101   1.5493317684      0
#> 102  -0.7733381025      0
#> 103   1.1077507292      0
#> 104   0.4482934995      0
#> 105   0.6791545871      0
#> 106   1.2910914279      0
#> 107  -0.0175577662      0
#> 108   0.6673978621      0
#> 109   0.5502857306      0
#> 110  -0.3641847641      0
#> 111   0.2068309154      0
#> 112   0.6457540657      0
#> 113  -0.5220890534      0
#> 114   1.6312031474      0
#> 115  -0.5519565749      0
#> 116   0.8566374146      0
#> 117  -0.4297096708      0
#> 118  -0.5642989710      0
#> 119   0.5062627011      0
#> 120   3.2370637929      0
#> 121  -1.2734236892      0
#> 122  -0.3738839853      0
#> 123  -0.4807453689      0
#> 124  -0.8024952721      0
#> 125  -0.8422588185      0
#> 126  -0.8424154773      0
#> 127   1.3838518473      0
#> 128  -0.1383420318      0
#> 129   0.2978521392      0
#> 130  -1.0068455076      0
#> 131  -0.8853909887      0
#> 132   0.2800419231      0
#> 133  -0.0871791112      0
#> 134   0.6065534720      0
#> 135  -0.3624662804      0
#> 136   0.4151749071      0
#> 137   0.2801609334      0
#> 138   0.0419047431      0
#> 139  -0.0861781501      0
#> 140   1.2372655488      0
#> 141   0.8205460012      0
#> 142   0.8143481957      0
#> 143   1.7742346618      0
#> 144   0.1452546466      0
#> 145   0.1343256873      0
#> 146  -1.1519424107      0
#> 147   0.0828407452      0
#> 148   0.1845094485      0
#> 149  -0.4170023054      0
#> 150  -0.4108378025      0
#> 151   0.3358386001      0
#> 152   0.3823799058      0
#> 153   1.5193298073      0
#> 154   0.9847803864      0
#> 155  -0.9683041058      0
#> 156  -1.4278349271      0
#> 157   1.0833943766      0
#> 158   0.6174772059      0
#> 159   0.8199357428      0
#> 160  -0.2616137777      0
#> 161  -0.2796492758      0
#> 162  -0.8065413174      0
#> 163  -0.9946341763      0
#> 164   0.8495279889      0
#> 165  -0.7053252532      0
#> 166  -1.1297806715      0
#> 167   0.4973452516      0
#> 168  -0.1166431950      0
#> 169  -0.2135687275      0
#> 170   0.3851534372      0
#> 171  -0.7801952032      0
#> 172   0.6389378744      0
#> 173  -0.1288544405      0
#> 174  -1.0560951222      0
#> 175  -0.6062311519      0
#> 176   1.0090313542      0
#> 177  -2.0574576462      0
#> 178  -0.4049907452      0
#> 179   0.0810405155      0
#> 180  -1.0553817216      0
#> 181   0.3467011710      0
#> 182   0.0529936861      0
#> 183  -0.6020918938      0
#> 184   0.5663273478      0
#> 185  -1.3241923140      0
#> 186   1.1318425382      0
#> 187   0.7291181280      0
#> 188  -1.2881627043      0
#> 189   0.3603644953      0
#> 190  -0.1173310298      0
#> 191  -0.6283028972      0
#> 192  -0.4752667761      0
#> 193   1.7423040312      0
#> 194  -0.5297339415      0
#> 195  -0.4908075646      0
#> 196   0.0817934851      0
#> 197  -0.6756853401      0
#> 198  -0.4175508744      0
#> 199  -0.6129196254      0
#> 200  -1.0662111307      0
#> 201   2.5981900964      0
#> 202  -0.5521355992      0
#> 203  -0.1514985428      0
#> 204  -1.4386993080      0
#> 205   0.2464800001      0
#> 206   1.7034655382      0
#> 207   0.6130219502      0
#> 208   0.8704625510      0
#> 209   0.3186877846      0
#> 210   0.1951801069      0
#> 211  -0.2055563025      0
#> 212   0.5743228422      0
#> 213  -1.1199317410      0
#> 214  -0.0032988366      0
#> 215   0.6427200790      0
#> 216   1.6749566643      0
#> 217   0.2396755819      0
#> 218  -0.4486709376      0
#> 219   0.5978018799      0
#> 220  -0.1351521966      0
#> 221   1.7494276046      0
#> 222   1.2727967045      0
#> 223   0.7336507288      0
#> 224   0.8692649427      0
#> 225  -1.6564211775      0
#> 226   0.9398421133      0
#> 227   0.5124399154      0
#> 228  -1.2756005998      0
#> 229   0.6671947633      0
#> 230   1.6156607446      0
#> 231  -0.2272558003      0
#> 232  -1.1616468878      0
#> 233   1.2156387109      0
#> 234   1.4095951094      0
#> 235   0.0234571542      0
#> 236   1.5400404307      0
#> 237   2.1393630936      0
#> 238  -0.7008052686      0
#> 239  -0.1297716908      0
#> 240  -2.2822093550      0
#> 241   0.3134017918      0
#> 242   0.2851523999      0
#> 243   1.1449279684      0
#> 244  -1.0500990883      0
#> 245  -0.0243998603      0
#> 246  -1.1236222569      0
#> 247  -0.4102376874      0
#> 248   1.2293846794      0
#> 249   0.1568051191      0
#> 250  -0.9396748692      0
#> 251   0.3594777249      0
#> 252  -0.5282684015      0
#> 253  -1.0430797271      0
#> 254   0.3389680066      0
#> 255  -0.2432807998      0
#> 256   0.8931921579      0
#> 257   0.6605606451      0
#> 258  -0.0258946192      0
#> 259   0.3467116377      0
#> 260   0.1910161183      0
#> 261  -0.7095450126      0
#> 262   0.0919821034      0
#> 263  -0.0169713481      0
#> 264   1.5020078118      0
#> 265   0.1889731998      0
#> 266  -0.1619502757      0
#> 267   0.4812984074      0
#> 268  -1.9789986261      0
#> 269   0.2742897280      0
#> 270  -0.8729413818      0
#> 271   0.9203952258      0
#> 272  -0.1039358813      0
#> 273  -0.2555731574      0
#> 274  -0.4382644426      0
#> 275   1.1526115396      0
#> 276  -0.7446233687      0
#> 277   0.8270421796      0
#> 278  -0.0780486721      0
#> 279   0.1810154712      0
#> 280   0.6797744665      0
#> 281   2.0074991801      0
#> 282  -0.0955900403      0
#> 283  -1.3818305667      0
#> 284  -1.4292489536      0
#> 285  -0.0809143454      0
#> 286   0.0966527378      0
#> 287  -1.0691408193      0
#> 288  -0.2637673266      0
#> 289   0.4469171602      0
#> 290   0.8364497446      0
#> 291   0.0327999307      0
#> 292   1.0186550304      0
#> 293  -1.2490042582      0
#> 294   0.8688784555      0
#> 295   0.6689937480      0
#> 296  -0.7131671406      0
#> 297  -0.2118477777      0
#> 298  -0.3252268020      0
#> 299   0.0190237890      0
#> 300  -2.1505381083      0
#> 301   1.2903850269      0
#> 302  -0.8615408182      0
#> 303   1.0680407630      0
#> 304  -1.2000898684      0
#> 305   0.6097667610      0
#> 306  -0.3547922814      0
#> 307   0.2276093024      0
#> 308   0.4805910654      0
#> 309  -0.4307163548      0
#> 310  -0.6442409265      0
#> 311   0.2780305609      0
#> 312  -1.0924575681      0
#> 313  -0.7485390218      0
#> 314   0.0712765689      0
#> 315  -0.8648837646      0
#> 316  -0.0706066713      0
#> 317  -0.2860047283      0
#> 318  -2.0521710163      0
#> 319  -0.8385527841      0
#> 320  -1.9419468687      0
#> 321   0.0994867744      0
#> 322   0.3559903719      0
#> 323   0.2582199515      0
#> 324   1.3686028448      0
#> 325  -0.9685922646      0
#> 326   0.8704575117      0
#> 327   0.0192361075      0
#> 328   0.1032894554      0
#> 329  -0.1171217403      0
#> 330  -0.2796838654      0
#> 331  -0.7405230860      0
#> 332  -0.7792753735      0
#> 333   0.3044810290      0
#> 334   0.2933859958      0
#> 335  -0.3678956157      0
#> 336  -0.0080781644      0
#> 337  -1.6643570252      0
#> 338  -0.3576833372      0
#> 339   0.9045441956      0
#> 340  -0.4204412652      0
#> 341   1.4781400271      0
#> 342  -1.8184711433      0
#> 343  -0.9862229882      0
#> 344   0.3372404945      0
#> 345  -1.8894572648      0
#> 346   0.0750036807      0
#> 347   1.1687246179      0
#> 348  -0.4722608284      0
#> 349  -0.2951827189      0
#> 350  -0.1318191709      0
#> 351   1.1648520185      0
#> 352   0.7214792268      0
#> 353   0.3332504300      0
#> 354   1.2716344136      0
#> 355  -0.3096101815      0
#> 356   0.9228835695      0
#> 357  -0.2400884289      0
#> 358   0.2752468875      0
#> 359  -0.5386253588      0
#> 360   0.1781823343      0
#> 361   1.7289112148      0
#> 362   0.2702829964      0
#> 363   1.3010977052      0
#> 364   0.6211035199      0
#> 365   0.4039192338      0
#> 366  -0.8024274961      0
#> 367  -0.0172489734      0
#> 368   0.0541096989      0
#> 369   0.2543870240      0
#> 370   0.9821439314      0
#> 371   0.0644460299      0
#> 372   2.4089639442      0
#> 373  -1.1935615930      0
#> 374  -1.3461592851      0
#> 375  -0.4011690036      0
#> 376  -0.6857950363      0
#> 377   0.5524430333      0
#> 378   0.1165003074      0
#> 379   0.3631652278      0
#> 380  -0.3423284200      0
#> 381   0.9048031194      0
#> 382  -0.4592866848      0
#> 383   0.1777843284      0
#> 384   2.9272691871      0
#> 385  -0.0002127017      0
#> 386   0.9699517900      0
#> 387   0.6152559001      0
#> 388  -0.4922063592      0
#> 389  -0.9371854209      0
#> 390  -0.5868420197      0
#> 391  -1.1884613944      0
#> 392   0.6126911042      0
#> 393  -0.2466274150      0
#> 394  -0.2609890232      0
#> 395  -0.2777443460      0
#> 396   0.1872827163      0
#> 397   1.4336994913      0
#> 398   0.8072650577      0
#> 399  -2.0267311546      0
#> 400  -0.4010114396      0
#> 401  -0.0629172638      0
#> 402  -0.1477856321      0
#> 403   0.3407247076      0
#> 404  -0.9298759098      0
#> 405  -0.5378249899      0
#> 406   1.2635335422      0
#> 407  -0.8354744839      0
#> 408   1.2680201195      0
#> 409  -0.3793953438      0
#> 410  -0.3103416234      0
#> 411   0.5159122347      0
#> 412  -0.3882528435      0
#> 413   0.4686842702      0
#> 414  -1.3872353245      0
#> 415   1.2364421774      0
#> 416  -1.3467407215      0
#> 417   0.4975045296      0
#> 418  -0.8596095598      0
#> 419   0.0597407724      0
#> 420   0.0217909390      0
#> 421  -0.3673496243      0
#> 422   1.5030098220      0
#> 423   0.4321454325      0
#> 424  -1.1401368093      0
#> 425  -0.6992914602      0
#> 426   1.4860157356      0
#> 427   1.9135253383      0
#> 428  -0.5452441965      0
#> 429   0.3123507516      0
#> 430  -0.1515020215      0
#> 431   1.2772201690      0
#> 432   0.0420180163      0
#> 433  -0.2556817492      0
#> 434   0.1906201345      0
#> 435  -0.5302813871      0
#> 436  -0.1389709482      0
#> 437  -1.2016216799      0
#> 438  -1.7041574080      0
#> 439   2.1096145495      0
#> 440   0.1814455879      0
#> 441   0.2974688404      0
#> 442   1.2370870876      0
#> 443   0.4882136428      0
#> 444  -0.4528318489      0
#> 445   0.2817990552      0
#> 446   0.4865767941      0
#> 447  -0.9796003692      0
#> 448  -1.6958015354      0
#> 449  -0.3622176534      0
#> 450  -0.0135179739      0
#> 451   0.6987930201      0
#> 452   0.7293864097      0
#> 453  -0.4241995017      0
#> 454   0.7351841145      0
#> 455  -0.0406361111      0
#> 456   0.2952661254      0
#> 457  -0.6490699684      0
#> 458   1.1929889931      0
#> 459  -1.8173012598      0
#> 460   0.4655121931      0
#> 461   0.8100210161      0
#> 462  -0.2165346854      0
#> 463  -0.1792935314      0
#> 464   2.4929134402      0
#> 465  -0.8357800109      0
#> 466  -1.6996168456      0
#> 467   0.0158077543      0
#> 468   1.3192106695      0
#> 469   0.0419010379      0
#> 470  -1.6004305983      0
#> 471  -0.1583930105      0
#> 472   0.5808836620      0
#> 473   2.3266812274      0
#> 474   1.4960229239      0
#> 475  -0.1949664549      0
#> 476   0.3776897168      0
#> 477  -0.3787644787      0
#> 478  -0.4125182651      0
#> 479   0.5863207073      0
#> 480  -0.1698887418      0
#> 481  -1.3862514854      0
#> 482   1.3804512601      0
#> 483  -1.5470711013      0
#> 484   0.2642924512      0
#> 485   0.7747408544      0
#> 486  -2.0591424726      0
#> 487  -1.5852744627      0
#> 488  -0.7806406592      0
#> 489   0.3206926206      0
#> 490  -2.6183925591      0
#> 491  -0.1709535310      0
#> 492  -0.8289720261      0
#> 493  -0.6132438155      0
#> 494   0.7550128680      0
#> 495   0.2229153293      0
#> 496   0.0801274923      0
#> 497  -0.0115943934      0
#> 498   0.0981121540      0
#> 499   0.0405599985      0
#> 500   0.7284343919      0
#> 501  -1.2765254045      0
#> 502  -2.1556807096      0
#> 503  -1.3028810261      0
#> 504  -0.0120092280      0
#> 505  -0.1706731294      0
#> 506   0.0175441902      0
#> 507  -2.4330788785      0
#> 508  -1.0859998678      0
#> 509   0.0652722220      0
#> 510   0.5371111235      0
#> 511   0.4147860234      0
#> 512   2.1953119025      0
#> 513   0.3556747732      0
#> 514  -0.8609483767      0
#> 515  -0.4584886654      0
#> 516  -1.1577068847      0
#> 517  -0.3777887837      0
#> 518   0.7386838082      0
#> 519  -0.0237850602      0
#> 520  -0.1432866469      0
#> 521  -1.0563909665      0
#> 522   1.2909571575      0
#> 523   0.4984931093      0
#> 524   1.4315009960      0
#> 525  -0.9330027287      0
#> 526   0.6137583223      0
#> 527   0.3166456979      0
#> 528  -0.2083171537      0
#> 529   0.3578884939      0
#> 530   0.1666852219      0
#> 531  -0.7850448669      0
#> 532   0.9802365727      0
#> 533   0.7864077380      0
#> 534   0.7169139718      0
#> 535   1.4052393866      0
#> 536  -1.3092177613      0
#> 537  -0.4565407334      0
#> 538   0.0811877415      0
#> 539   0.8494365951      0
#> 540  -1.0112620625      0
#> 541  -1.1440329297      0
#> 542   1.7378711718      0
#> 543  -1.8774106272      0
#> 544   1.3039601683      0
#> 545   0.0811197911      0
#> 546  -0.7625592533      0
#> 547  -0.3414424036      0
#> 548   0.4331806761      0
#> 549   1.4765676790      0
#> 550  -0.9998249484      0
#> 551  -0.0312444927      0
#> 552   0.3064057965      0
#> 553  -0.0826064849      0
#> 554  -0.8578747573      0
#> 555  -0.2443721923      0
#> 556   0.4016344151      0
#> 557  -1.1646101096      0
#> 558   1.8635178126      0
#> 559  -0.6851958275      0
#> 560   0.2691003123      0
#> 561   1.8373918239      0
#> 562  -1.6814542886      0
#> 563   0.2121525162      0
#> 564   1.2984626083      0
#> 565   0.5622582853      0
#> 566   0.4810957528      0
#> 567  -0.1550079777      0
#> 568   0.9181873725      0
#> 569   0.7875200993      0
#> 570   0.2596472243      0
#> 571  -0.3981565981      0
#> 572   0.6682211230      0
#> 573  -1.1062686532      0
#> 574  -0.1528751503      0
#> 575   1.1424509849      0
#> 576   0.4633819542      0
#> 577  -1.0811129021      0
#> 578  -1.2710484892      0
#> 579   0.0490464803      0
#> 580   0.4968103003      0
#> 581  -1.3581658317      0
#> 582  -1.6589048633      0
#> 583  -0.5787522695      0
#> 584  -0.4910559067      0
#> 585  -0.0642749108      0
#> 586  -0.7077836844      0
#> 587  -1.6220283402      0
#> 588   1.5702110531      0
#> 589  -0.2733309320      0
#> 590   0.5318246613      0
#> 591  -0.0794246494      0
#> 592   0.2915044955      0
#> 593   0.1102875736      0
#> 594   0.1743384941      0
#> 595  -1.0372867904      0
#> 596  -1.3697178056      0
#> 597  -0.3728950973      0
#> 598   0.6708283906      0
#> 599  -2.0713934858      0
#> 600   0.2803451892      0
#> 601  -0.0195331581      0
#> 602  -1.3792585789      0
#> 603  -0.4013844152      0
#> 604  -1.7849693373      0
#> 605  -0.9423193187      0
#> 606   0.5454482445      0
#> 607   0.2383884067      0
#> 608   1.1185033695      0
#> 609  -0.2215689662      0
#> 610  -1.8623446497      0
#> 611   0.1901380382      0
#> 612  -0.8051937704      0
#> 613  -1.7573745792      0
#> 614  -0.8991605406      0
#> 615  -1.2050876412      0
#> 616   0.2643944801      0
#> 617  -0.2591737719      0
#> 618  -0.6971422268      0
#> 619  -1.1933498062      0
#> 620  -2.9490401577      0
#> 621  -0.1424140966      0
#> 622   0.7082759221      0
#> 623   0.6997487491      0
#> 624  -0.4113437780      0
#> 625   0.8437124117      0
#> 626   2.3943030250      0
#> 627   0.5130413138      0
#> 628   1.5566157812      0
#> 629  -0.0366356020      0
#> 630  -0.6597062489      0
#> 631   0.6950529726      0
#> 632  -0.1926672567      0
#> 633   1.0173925419      0
#> 634   1.8782959670      0
#> 635  -0.8995163892      0
#> 636  -0.6212242969      0
#> 637   0.3840913872      0
#> 638   1.3229744242      0
#> 639   0.1557323936      0
#> 640  -0.3411833164      0
#> 641   0.8328438428      0
#> 642   0.1479209836      0
#> 643   1.7402859710      0
#> 644  -0.0383960797      0
#> 645   0.3640161290      0
#> 646  -0.2500949071      0
#> 647   1.2081506906      0
#> 648   1.3721149445      0
#> 649  -0.7495299912      0
#> 650  -0.2292118360      0
#> 651   1.7827912119      0
#> 652   0.7135262117      0
#> 653  -0.5191979816      0
#> 654  -0.5560130628      0
#> 655  -1.0409977828      0
#> 656  -0.5171194755      0
#> 657  -1.9103102317      0
#> 658  -0.5395926788      0
#> 659  -1.6070243220      0
#> 660  -0.8733395318      0
#> 661   0.6041753962      0
#> 662   0.2320704254      0
#> 663   1.1543600541      0
#> 664   0.5012722602      0
#> 665  -0.0646221904      0
#> 666  -0.1604002308      0
#> 667  -0.4850716190      0
#> 668  -0.4243086337      0
#> 669   0.3488195463      0
#> 670  -1.6024047255      0
#> 671  -0.8160903840      0
#> 672  -3.3934940643      0
#> 673  -0.2047795943      0
#> 674   0.6123660725      0
#> 675  -0.6636288606      0
#> 676  -0.5636847818      0
#> 677  -0.6829219229      0
#> 678   0.2171429769      0
#> 679  -1.6268621632      0
#> 680   0.8786736748      0
#> 681  -0.9741299760      0
#> 682   0.6602922650      0
#> 683   1.0782524877      0
#> 684   0.1575169073      0
#> 685   0.5520004502      0
#> 686   0.2460606643      0
#> 687   0.0949063472      0
#> 688   0.7478001474      0
#> 689   0.0404209430      0
#> 690  -0.1760805467      0
#> 691  -1.3878571697      0
#> 692   0.6041302418      0
#> 693  -0.3563888411      0
#> 694   0.9752490105      0
#> 695  -1.1652481874      0
#> 696  -0.3657692413      0
#> 697   0.4615897127      0
#> 698   0.7647764105      0
#> 699   1.0867658360      0
#> 700   0.0420596367      0
#> 701   0.2419276750      0
#> 702   1.6074450494      0
#> 703   1.3798531045      0
#> 704   1.0159865529      0
#> 705  -1.4822212750      0
#> 706   0.5056760131      0
#> 707   1.2373207167      0
#> 708  -1.9998051346      0
#> 709  -0.8844967638      0
#> 710   1.7239838058      0
#> 711  -1.3411058984      0
#> 712  -0.0730410735      0
#> 713  -0.0222972779      0
#> 714   1.5539894536      0
#> 715  -1.8643758100      0
#> 716  -0.0311064667      0
#> 717  -2.2572089201      0
#> 718  -1.7362458411      0
#> 719  -0.8329789525      0
#> 720   1.8801288676      0
#> 721   0.8763763460      0
#> 722   0.0402516901      0
#> 723  -2.0864822678      0
#> 724  -1.0370257827      0
#> 725   1.0599083446      0
#> 726  -0.0159954258      0
#> 727   0.0326513546      0
#> 728  -0.3567049052      0
#> 729   0.7407414030      0
#> 730   1.2991071575      0
#> 731  -0.4259732610      0
#> 732   0.6728992022      0
#> 733  -0.8775119920      0
#> 734   0.3655906673      0
#> 735   0.9914995285      0
#> 736   0.9493383439      0
#> 737   0.0382567773      0
#> 738  -1.0559811562      0
#> 739   0.2993974455      0
#> 740  -0.3387289958      0
#> 741   0.6550530751      0
#> 742  -0.5404044387      0
#> 743  -1.4494164010      0
#> 744  -1.4105671583      0
#> 745   1.0541666532      0
#> 746  -0.0974820239      0
#> 747  -0.5498574509      0
#> 748  -0.7311234547      0
#> 749  -0.5042349495      0
#> 750  -0.7786956585      0
#> 751   1.9442920395      0
#> 752   0.1215961313      0
#> 753  -2.0790796081      0
#> 754  -1.1030581990      0
#> 755  -0.8327324001      0
#> 756   0.1525857013      0
#> 757   0.0529197914      0
#> 758  -0.7353281163      0
#> 759  -0.2665131724      0
#> 760   0.2832567994      0
#> 761  -1.2329780886      0
#> 762   0.1687998671      0
#> 763  -0.5537378391      0
#> 764  -0.6024815890      0
#> 765  -1.2965196340      0
#> 766   0.4224781737      0
#> 767   0.0723131193      0
#> 768   0.0268818396      0
#> 769   0.4054968818      0
#> 770   0.6817871590      0
#> 771  -0.5611045253      0
#> 772  -0.4850517977      0
#> 773  -1.3548600037      0
#> 774   1.1581313615      0
#> 775   2.6107150090      0
#> 776   0.0378287502      0
#> 777  -0.3211554273      0
#> 778  -1.1818347822      0
#> 779  -0.5403596956      0
#> 780   0.9347393198      0
#> 781  -0.3888452427      0
#> 782   0.5464641090      0
#> 783  -0.1100894428      0
#> 784  -0.4731789384      0
#> 785   0.8043258574      0
#> 786   1.3527955121      0
#> 787   0.0227578328      0
#> 788   0.6928572564      0
#> 789  -0.0550005606      0
#> 790  -1.1020687528      0
#> 791  -0.6473544802      0
#> 792   2.4205587279      0
#> 793   0.9184769828      0
#> 794  -1.4030687542      0
#> 795  -0.3095892527      0
#> 796  -0.6463252039      0
#> 797  -0.0782485831      0
#> 798   0.4604091454      0
#> 799  -0.9761280679      0
#> 800  -0.2163849805      0
#> 801  -0.1744434013      0
#> 802  -0.2089354818      0
#> 803  -0.0002304708      0
#> 804   0.2178123953      0
#> 805   0.2234907643      0
#> 806  -1.3071238995      0
#> 807   0.4666392714      0
#> 808   0.4994813321      0
#> 809   0.9968904777      0
#> 810  -0.4880450750      0
#> 811  -0.5472152669      0
#> 812  -0.2293979382      0
#> 813   0.6143718029      0
#> 814  -0.4614955336      0
#> 815  -0.2934748429      0
#> 816  -0.1230442924      0
#> 817  -1.3210732563      0
#> 818  -0.8736921224      0
#> 819   0.6575334635      0
#> 820   0.0534857769      0
#> 821  -0.5529145681      0
#> 822  -0.2228450462      0
#> 823   0.1187835793      0
#> 824   0.6403183221      0
#> 825  -1.1037487674      0
#> 826  -0.4674897633      0
#> 827  -0.8296321815      0
#> 828  -1.5118629178      0
#> 829  -0.2646530942      0
#> 830  -0.0010311242      0
#> 831   1.1466396842      0
#> 832  -2.8143967672      0
#> 833  -0.2846184166      0
#> 834   0.5108311021      0
#> 835   0.9453623744      0
#> 836   0.0950215267      0
#> 837  -0.4985637264      0
#> 838  -1.1814955630      0
#> 839   0.3242903810      0
#> 840   0.1102914059      0
#> 841   0.1796421721      0
#> 842   0.6852213722      0
#> 843  -0.1108933381      0
#> 844  -0.7655111049      0
#> 845   1.3160503815      0
#> 846   0.2553693004      0
#> 847  -0.8198190161      0
#> 848   0.7279445007      0
#> 849  -0.1792270811      0
#> 850  -0.1513778564      0
#> 851  -0.8184274266      0
#> 852   1.9910354718      0
#> 853  -0.1318207990      0
#> 854  -2.9227645990      0
#> 855   1.6937699744      0
#> 856  -0.0828261826      0
#> 857   0.0699016602      0
#> 858   0.7599839164      0
#> 859  -1.0547840030      0
#> 860  -1.5783479164      0
#> 861   2.1504669211      0
#> 862  -1.5128467419      0
#> 863  -0.9456245167      0
#> 864  -0.2990302728      0
#> 865   0.5391164618      0
#> 866   0.3801604065      0
#> 867  -0.8170443622      0
#> 868   1.1476465066      0
#> 869   0.5950272974      0
#> 870   0.1546740371      0
#> 871   1.0097547205      0
#> 872  -1.4608828711      0
#> 873   0.0297604529      0
#> 874  -0.4976888033      0
#> 875   0.2849411234      0
#> 876  -0.0433677343      0
#> 877   1.0175367893      0
#> 878   0.7877912789      0
#> 879   1.5543773349      0
#> 880  -0.2391513053      0
#> 881  -1.1360073841      0
#> 882  -0.5776465965      0
#> 883  -1.8168726184      0
#> 884  -0.1170381873      0
#> 885  -0.6837493009      0
#> 886   1.1546356163      0
#> 887   0.9243278272      0
#> 888  -0.1956581839      0
#> 889   1.1078953267      0
#> 890  -0.1961626966      0
#> 891   1.6966552125      0
#> 892  -0.5640094326      0
#> 893  -0.4319113202      0
#> 894   0.2475060520      0
#> 895  -1.0917953321      0
#> 896   0.6113855254      0
#> 897  -0.7109333280      0
#> 898  -0.7102057523      0
#> 899  -0.4098435775      0
#> 900   0.1996337158      0
#> 901   0.8121592524      0
#> 902   0.5324736770      0
#> 903   1.6439787965      0
#> 904   0.6638571403      0
#> 905   1.1692225817      0
#> 906  -2.1807150976      0
#> 907  -0.1133411132      0
#> 908   2.3584768130      0
#> 909   1.5481642341      0
#> 910  -1.0445485977      0
#> 911  -0.8551242435      0
#> 912   0.2118693389      0
#> 913   1.0439503326      0
#> 914  -1.1327058757      0
#> 915   1.4523600375      1
#> 916   0.7200287484      1
#> 917   1.9748232633      1
#> 918  -1.1204556109      1
#> 919  -1.4091195594      1
#> 920   1.5026010846      1
#> 921   1.5699547799      1
#> 922  -0.1332082236      1
#> 923   2.4433737155      1
#> 924   0.5786669166      1
#> 925  -0.1928722217      1
#> 926  -0.1616710290      1
#> 927   0.7990937235      1
#> 928   0.8542861007      1
#> 929   0.8846853998      1
#> 930   1.5529447480      1
#> 931   0.6416905644      1
#> 932   0.2723342390      1
#> 933   0.0959311324      1
#> 934   0.5035740413      1
#> 935  -0.4387268765      1
#> 936   0.1057925410      1
#> 937   0.8953102598      1
#> 938  -0.5557992202      1
#> 939   0.6593791691      1
#> 940  -0.0387908902      1
#> 941   1.6053559348      1
#> 942   2.7915539099      1
#> 943   0.7619156945      1
#> 944   1.4972041894      1
#> 945   0.7861058122      1
#> 946   0.1013238431      1
#> 947  -0.0126593330      1
#> 948   0.6273815568      1
#> 949   1.0748376849      1
#> 950   0.9102982006      1
#> 951   0.8458179763      1
#> 952   0.1192177389      1
#> 953  -0.0009496419      1
#> 954   1.9036345182      1
#> 955   0.7824761796      1
#> 956   0.8757286489      1
#> 957   3.2892549984      1
#> 958   1.3268026128      1
#> 959   0.1708485450      1
#> 960   1.8096537549      1
#> 961   0.9223282446      1
#> 962   0.8827129201      1
#> 963   3.1958550310      1
#> 964   1.2287223298      1
#> 965  -0.6985788784      1
#> 966   1.3135894469      1
#> 967   0.9297938233      1
#> 968  -0.6450462953      1
#> 969   1.7872279993      1
#> 970   0.2664619693      1
#> 971   0.7742337131      1
#> 972   0.2451332378      1
#> 973   0.7857831324      1
#> 974   1.5744915650      1
#> 975   0.5512685222      1
#> 976  -0.0593212168      1
#> 977   1.4985073554      1
#> 978  -0.1979726593      1
#> 979   1.8460004612      1
#> 980   1.9058398578      1
#> 981   0.1180741516      1
#> 982   1.5198995534      1
#> 983   2.0960429037      1
#> 984   0.6292267605      1
#> 985   0.8663789041      1
#> 986  -0.2999685662      1
#> 987   0.4812738830      1
#> 988   0.8799840827      1
#> 989   0.1968720298      1
#> 990   2.0777936994      1
#> 991   0.5302302644      1
#> 992  -0.7087746826      1
#> 993   1.6001758962      1
#> 994   1.1122617802      1
#> 995  -0.4564641539      1
#> 996   3.1625060607      1
#> 997   0.7205378710      1
#> 998   1.6395893796      1
#> 999   0.8905202799      1
#> 1000 -0.5632916269      1