Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     0.6599699651      0
#> 2     0.5587094556      0
#> 3    -0.5374771726      0
#> 4    -1.5996959729      0
#> 5    -0.6032427136      0
#> 6    -1.1736540107      0
#> 7    -0.8689100708      0
#> 8    -0.3831180252      0
#> 9     0.4891492951      0
#> 10   -0.2172203273      0
#> 11    1.5080013152      0
#> 12   -0.9130445095      0
#> 13    0.8531345681      0
#> 14    0.0665349679      0
#> 15   -1.0217299226      0
#> 16   -0.8249451703      0
#> 17    0.2977223657      0
#> 18    0.0867590828      0
#> 19   -1.0570927815      0
#> 20    0.9346103269      0
#> 21   -0.1457313179      0
#> 22    0.4126739216      0
#> 23    1.0266617238      0
#> 24    1.1255594673      0
#> 25   -1.6225942737      0
#> 26    0.3046476085      0
#> 27   -0.5021635623      0
#> 28    0.1641107065      0
#> 29    1.5970148505      0
#> 30   -0.3648184729      0
#> 31    1.0705320346      0
#> 32    0.3590295058      0
#> 33   -0.3314656517      0
#> 34    2.1133762472      0
#> 35    1.0752487343      0
#> 36   -1.2922343786      0
#> 37   -1.3487288245      0
#> 38    0.1111981091      0
#> 39   -0.0761901900      0
#> 40    0.2713546306      0
#> 41    0.0152035963      0
#> 42    0.2695896602      0
#> 43    1.1140418461      0
#> 44   -0.7927113713      0
#> 45   -1.7572673488      0
#> 46   -2.0836481786      0
#> 47   -1.1980496360      0
#> 48    1.3348908734      0
#> 49    0.7296190117      0
#> 50   -0.5954473624      0
#> 51   -0.6807022848      0
#> 52    1.6334654400      0
#> 53    0.5097395038      0
#> 54    0.0355922024      0
#> 55   -1.3317368369      0
#> 56   -0.5141234021      0
#> 57    0.0816491537      0
#> 58   -0.4705289677      0
#> 59   -0.3850781997      0
#> 60    0.3276110847      0
#> 61    1.5511931110      0
#> 62    0.1416103739      0
#> 63   -0.8614907431      0
#> 64   -0.2902278406      0
#> 65    1.4534469336      0
#> 66   -0.3492310191      0
#> 67    0.5039085181      0
#> 68    1.8159857868      0
#> 69   -1.7981240885      0
#> 70    1.9902675281      0
#> 71    1.7893990438      0
#> 72   -0.3024070755      0
#> 73   -0.4910485769      0
#> 74   -0.3545394238      0
#> 75    0.4661468538      0
#> 76    0.4647124026      0
#> 77    0.0680741129      0
#> 78   -0.0667659920      0
#> 79    1.0461052378      0
#> 80    1.5520274919      0
#> 81    1.5311066325      0
#> 82    1.5360988749      0
#> 83    0.0119089751      0
#> 84    1.3113819039      0
#> 85   -1.7900839220      0
#> 86   -0.6741851662      0
#> 87   -1.3324105506      0
#> 88    1.0992771186      0
#> 89    0.6048353106      0
#> 90    1.2779770744      0
#> 91   -0.4610192143      0
#> 92   -0.8622288328      0
#> 93   -0.7981266424      0
#> 94   -0.6701009289      0
#> 95   -0.9034564019      0
#> 96    0.2708572481      0
#> 97    0.9436029743      0
#> 98   -0.0530700270      0
#> 99    1.1685038924      0
#> 100  -0.7182462525      0
#> 101  -0.5197966319      0
#> 102   0.8378478063      0
#> 103   0.9487609970      0
#> 104  -0.2278014703      0
#> 105  -1.6525503051      0
#> 106  -0.9557352258      0
#> 107  -0.6999137212      0
#> 108  -0.6872025229      0
#> 109   0.7782907391      0
#> 110  -1.0517009037      0
#> 111   0.9768260800      0
#> 112  -0.4671518748      0
#> 113   0.9743488406      0
#> 114  -0.9423183058      0
#> 115   1.5128765485      0
#> 116   0.2503691206      0
#> 117   0.3734570472      0
#> 118   0.1207816854      0
#> 119   0.5683500199      0
#> 120   0.7808316471      0
#> 121  -1.2511152321      0
#> 122   0.3206446060      0
#> 123   1.8599454588      0
#> 124  -0.9860266422      0
#> 125  -0.8915998855      0
#> 126  -1.4018097117      0
#> 127  -0.0234419224      0
#> 128  -0.8152925677      0
#> 129  -0.0947346641      0
#> 130   1.2916997674      0
#> 131  -1.3707796222      0
#> 132   0.1752799350      0
#> 133   0.7595597281      0
#> 134   1.6902426085      0
#> 135  -0.7741822381      0
#> 136  -0.7691033468      0
#> 137   0.8001009493      0
#> 138  -0.4014496185      0
#> 139  -0.6052023533      0
#> 140  -0.7074323779      0
#> 141  -0.1371763353      0
#> 142   0.3623014839      0
#> 143  -0.5079837411      0
#> 144   0.5176179111      0
#> 145  -1.2194575799      0
#> 146  -1.3168988672      0
#> 147  -0.8521476717      0
#> 148   0.7619498916      0
#> 149   0.0047987462      0
#> 150   0.3200439026      0
#> 151   0.0584320197      0
#> 152  -0.6314821687      0
#> 153   0.7603366683      0
#> 154   0.6896706402      0
#> 155  -0.0984880889      0
#> 156  -0.4459645218      0
#> 157   0.1727850153      0
#> 158   1.6009942479      0
#> 159   0.1231237868      0
#> 160   1.1577348319      0
#> 161  -0.2278976455      0
#> 162   0.2525937254      0
#> 163  -0.9705470326      0
#> 164   0.7474027043      0
#> 165   0.5455491457      0
#> 166   1.1582103371      0
#> 167  -0.3982402618      0
#> 168  -0.4812439028      0
#> 169  -0.4577652631      0
#> 170  -0.3379138147      0
#> 171   1.3286502650      0
#> 172  -0.3790757266      0
#> 173  -0.0152318377      0
#> 174  -0.9966323350      0
#> 175   0.4720022288      0
#> 176   1.7048901353      0
#> 177  -0.9414200825      0
#> 178  -0.0939417309      0
#> 179   1.2021407969      0
#> 180   0.4447523077      0
#> 181  -0.8449410444      0
#> 182  -1.0379285638      0
#> 183  -1.4981667285      0
#> 184  -1.0746124026      0
#> 185  -0.0114612196      0
#> 186  -0.1365356545      0
#> 187  -0.3222257487      0
#> 188   1.1804654593      0
#> 189  -0.7362584728      0
#> 190   1.7696672594      0
#> 191  -1.5066381136      0
#> 192  -1.3346043719      0
#> 193   0.0541329325      0
#> 194   0.4596853565      0
#> 195  -0.4158286608      0
#> 196   0.6609151557      0
#> 197  -0.6048727373      0
#> 198   1.3488041107      0
#> 199  -0.9685999616      0
#> 200   0.3912526775      0
#> 201   0.3005537750      0
#> 202  -0.6481332592      0
#> 203  -0.2835601215      0
#> 204  -0.6025197629      0
#> 205   0.6981230157      0
#> 206   0.2563619420      0
#> 207   0.0418593375      0
#> 208  -0.1373918570      0
#> 209  -0.3192650668      0
#> 210  -0.7224998724      0
#> 211   1.3740994082      0
#> 212   0.1861355318      0
#> 213  -1.5107885467      0
#> 214   0.1206783001      0
#> 215  -0.6855582763      0
#> 216  -0.0029983256      0
#> 217  -0.1042576805      0
#> 218   0.8820421912      0
#> 219  -0.6632803397      0
#> 220   1.0726603356      0
#> 221   1.2532995856      0
#> 222   1.6754283671      0
#> 223   1.1911867021      0
#> 224   0.6968920167      0
#> 225  -0.2978277909      0
#> 226  -0.7798576127      0
#> 227  -1.1072882509      0
#> 228  -1.3575893652      0
#> 229  -0.0606724730      0
#> 230  -0.3527181044      0
#> 231   0.6100612451      0
#> 232   0.0963691651      0
#> 233  -0.2833878497      0
#> 234  -0.9122646613      0
#> 235  -0.0215315316      0
#> 236   2.2870263437      0
#> 237   0.5292409644      0
#> 238  -0.1898590393      0
#> 239   1.1546232539      0
#> 240  -0.0539524080      0
#> 241   0.9288426071      0
#> 242   0.6476950907      0
#> 243   0.8802495638      0
#> 244  -1.1365089563      0
#> 245   0.6944233095      0
#> 246   1.9214430981      0
#> 247  -0.0423796967      0
#> 248  -0.0057403091      0
#> 249   1.3831515554      0
#> 250  -0.3151773909      0
#> 251   0.3287519652      0
#> 252   1.2173913151      0
#> 253   0.3401955616      0
#> 254  -0.6983423575      0
#> 255  -1.6763716157      0
#> 256  -1.2880190155      0
#> 257   1.9384159375      0
#> 258   0.0616180597      0
#> 259   1.0276874608      0
#> 260   0.2915039064      0
#> 261  -0.4612869481      0
#> 262   0.0868267881      0
#> 263  -0.3185569612      0
#> 264  -1.3254587247      0
#> 265   1.5555986409      0
#> 266   0.6376679024      0
#> 267   1.4593112790      0
#> 268  -0.4649854674      0
#> 269  -2.1039012445      0
#> 270  -0.6462001668      0
#> 271  -0.4066602897      0
#> 272   0.8161288874      0
#> 273  -0.5999840445      0
#> 274   2.0842684962      0
#> 275   1.3551774693      0
#> 276  -1.3717862708      0
#> 277  -1.5325142411      0
#> 278  -0.6092202958      0
#> 279  -0.4285388230      0
#> 280   0.7250747550      0
#> 281   2.2463428266      0
#> 282  -1.4779367065      0
#> 283   0.5424128257      0
#> 284   0.4847184723      0
#> 285   0.8944702760      0
#> 286  -0.8849040643      0
#> 287   2.1015006500      0
#> 288  -0.7956611135      0
#> 289  -1.8672220375      0
#> 290  -0.8710715127      0
#> 291   1.0331611402      0
#> 292   1.3671007650      0
#> 293   0.0601558556      0
#> 294  -0.7382198086      0
#> 295   0.4033688231      0
#> 296  -0.5436993162      0
#> 297   0.8384420154      0
#> 298  -0.5976385697      0
#> 299   0.7702838112      0
#> 300   0.2821749332      0
#> 301  -0.3070717426      0
#> 302   0.0417417055      0
#> 303  -2.5968203922      0
#> 304  -0.5641513120      0
#> 305   1.2905161528      0
#> 306   0.8349873827      0
#> 307   0.5783378905      0
#> 308   1.0379618259      0
#> 309   0.7842556013      0
#> 310  -1.0139880812      0
#> 311   0.8514955647      0
#> 312  -0.5049240724      0
#> 313  -1.7826247632      0
#> 314  -0.1924851993      0
#> 315  -0.7225189589      0
#> 316  -2.1173759369      0
#> 317  -0.8006546423      0
#> 318  -0.6122235024      0
#> 319   0.1615319358      0
#> 320   0.9511669908      0
#> 321  -0.7298427124      0
#> 322   0.6623061820      0
#> 323  -0.0476737846      0
#> 324   1.5281281062      0
#> 325  -1.4431989570      0
#> 326   2.0882018278      0
#> 327   0.5557836760      0
#> 328   0.7893898338      0
#> 329   0.0093434064      0
#> 330   0.7706092651      0
#> 331  -1.2180864846      0
#> 332  -0.0309569023      0
#> 333  -0.2131092133      0
#> 334   1.3094586756      0
#> 335  -0.7699745426      0
#> 336   0.4940559115      0
#> 337   0.8114415490      0
#> 338  -1.0550275946      0
#> 339  -0.6218648136      0
#> 340   1.9155862101      0
#> 341  -2.1781592289      0
#> 342   0.3181256944      0
#> 343  -1.7064723538      0
#> 344  -0.9924026996      0
#> 345  -1.0733826944      0
#> 346  -0.5162351150      0
#> 347   1.6214503187      0
#> 348   0.5368980215      0
#> 349  -1.0701809231      0
#> 350  -0.8085423418      0
#> 351   0.2909624867      0
#> 352  -0.1720531084      0
#> 353  -0.4730289347      0
#> 354   1.2784868154      0
#> 355  -1.2603797922      0
#> 356  -0.7478540655      0
#> 357   0.3114821243      0
#> 358  -0.9420696840      0
#> 359  -0.1138230610      0
#> 360   1.5872791537      0
#> 361  -0.5434637296      0
#> 362  -0.3956081468      0
#> 363   1.4182312742      0
#> 364   1.2415348880      0
#> 365   0.1849453017      0
#> 366   0.1101346673      0
#> 367  -0.9444992656      0
#> 368   0.5787466084      0
#> 369  -0.7291323004      0
#> 370   0.1382860756      0
#> 371  -1.1797041365      0
#> 372  -0.3176387845      0
#> 373  -0.3644086535      0
#> 374  -0.1771683359      0
#> 375   0.2204978668      0
#> 376  -0.7787992505      0
#> 377   0.6774296303      0
#> 378  -0.5666247330      0
#> 379   0.4854238919      0
#> 380   1.2350874239      0
#> 381   1.0602217541      0
#> 382  -0.4068109850      0
#> 383   2.2156631004      0
#> 384  -0.1006102880      0
#> 385  -1.5795896934      0
#> 386   0.0588719134      0
#> 387  -0.0460285848      0
#> 388   1.6410803610      0
#> 389   0.3986329113      0
#> 390   0.2439657556      0
#> 391  -0.1547809263      0
#> 392   2.3162200639      0
#> 393   0.9309659846      0
#> 394   0.4504091207      0
#> 395   1.8157435690      0
#> 396  -0.3847073020      0
#> 397  -0.9630120334      0
#> 398  -0.4723282988      0
#> 399   0.1926582781      0
#> 400  -0.3724560816      0
#> 401   0.5833317871      0
#> 402   0.0559275435      0
#> 403   0.3092213632      0
#> 404   1.5460119213      0
#> 405  -0.0299976946      0
#> 406  -0.7597118184      0
#> 407  -0.3764285178      0
#> 408   0.2688191545      0
#> 409   2.4396628034      0
#> 410   1.3056797891      0
#> 411  -1.0559929592      0
#> 412   1.0155817516      0
#> 413   0.7206911025      0
#> 414  -0.6077515673      0
#> 415  -1.3251871891      0
#> 416   0.5113522282      0
#> 417   0.0990549429      0
#> 418   1.7663699698      0
#> 419   0.8618446527      0
#> 420   2.0481198786      0
#> 421   1.1213641766      0
#> 422   0.7529157343      0
#> 423  -0.3537491880      0
#> 424   1.4482291753      0
#> 425   0.9904558618      0
#> 426   0.3247653629      0
#> 427  -0.2863751504      0
#> 428   0.0198994205      0
#> 429  -2.0018208057      0
#> 430  -0.3710818205      0
#> 431  -0.4157603510      0
#> 432   0.7792470091      0
#> 433  -1.3161483194      0
#> 434  -0.1952080752      0
#> 435   0.4914372475      0
#> 436   0.2943319608      0
#> 437   0.8273786368      0
#> 438   1.6531489100      0
#> 439   0.2299353478      0
#> 440   0.5775842531      0
#> 441  -0.0880096429      0
#> 442  -0.3805896887      0
#> 443   0.5758239226      0
#> 444  -0.8390400823      0
#> 445   0.1657275433      0
#> 446  -1.3926991766      0
#> 447  -0.6547044387      0
#> 448  -0.3943156864      0
#> 449  -0.6163216416      0
#> 450   1.0316726859      0
#> 451  -0.9875160233      0
#> 452   1.1722601388      0
#> 453  -1.4016244093      0
#> 454   0.0381487890      0
#> 455   0.2337460993      0
#> 456  -0.2890886066      0
#> 457   0.3366829880      0
#> 458  -1.6267326310      0
#> 459  -1.4900289958      0
#> 460  -0.8444172955      0
#> 461   1.1984343599      0
#> 462   1.8542477601      0
#> 463   0.4802828401      0
#> 464   1.1371424740      0
#> 465  -0.0104095228      0
#> 466  -0.7509439936      0
#> 467   0.1765646601      0
#> 468  -0.1055788578      0
#> 469   0.8484711903      0
#> 470   1.0761169206      0
#> 471  -0.7714663701      0
#> 472  -0.0241223323      0
#> 473   2.4360696217      0
#> 474  -0.2496435497      0
#> 475   0.7523797207      0
#> 476  -0.3361812298      0
#> 477  -1.7479953058      0
#> 478   0.0329725820      0
#> 479  -0.3586844532      0
#> 480   0.4707802196      0
#> 481   1.2061418101      0
#> 482  -2.3452778052      0
#> 483  -0.8334911972      0
#> 484  -0.4732145903      0
#> 485   0.2567458373      0
#> 486  -0.1845342017      0
#> 487  -1.6554973062      0
#> 488  -0.1713197022      0
#> 489   2.1043139740      0
#> 490   2.0688229204      0
#> 491   1.3589083572      0
#> 492   0.7537413982      0
#> 493   0.5812249651      0
#> 494   0.7407290581      0
#> 495   0.6390400575      0
#> 496   0.5736427482      0
#> 497   0.4869683665      0
#> 498  -0.1289974776      0
#> 499  -1.7077512999      0
#> 500  -0.2026332407      0
#> 501  -1.5657315681      0
#> 502   1.3766853204      0
#> 503   1.5634779940      0
#> 504   0.1304064553      0
#> 505  -0.8035165570      0
#> 506   0.7653009429      0
#> 507   0.2203125321      0
#> 508   0.7501569854      0
#> 509   1.9945916860      0
#> 510  -0.6410604491      0
#> 511  -0.2827115044      0
#> 512   2.5275665700      0
#> 513  -0.3225938755      0
#> 514  -0.9287502146      0
#> 515   0.7982635257      0
#> 516  -1.1127108309      0
#> 517   1.0898644331      0
#> 518  -0.1274357132      0
#> 519  -0.8674432105      0
#> 520  -0.3463327615      0
#> 521   0.1199717111      0
#> 522   0.6164417289      0
#> 523   0.6168658196      0
#> 524  -1.3401716805      0
#> 525   0.4688659334      0
#> 526  -0.0277320900      0
#> 527   0.5618095674      0
#> 528   0.3812381898      0
#> 529  -0.3317243556      0
#> 530   0.5080289175      0
#> 531  -0.4983467642      0
#> 532   1.0855050826      0
#> 533   0.3098283547      0
#> 534   0.3770655920      0
#> 535   0.1503551053      0
#> 536  -0.7568619918      0
#> 537  -0.4790182085      0
#> 538  -0.0384998935      0
#> 539  -1.3108133195      0
#> 540  -0.7285831551      0
#> 541  -0.1179141356      0
#> 542   0.5832366830      0
#> 543   1.5778571303      0
#> 544  -0.6235211261      0
#> 545  -1.6215241890      0
#> 546   2.4897710392      0
#> 547   0.5811766340      0
#> 548   0.0517663221      0
#> 549  -0.2214932874      0
#> 550   1.1605141877      0
#> 551   1.2419763740      0
#> 552   0.5854202625      0
#> 553   0.4088296389      0
#> 554  -1.9099399519      0
#> 555  -0.4752956860      0
#> 556   0.9909698930      0
#> 557  -0.4979277110      0
#> 558   0.2090115585      0
#> 559   0.2642521766      0
#> 560  -0.9058923250      0
#> 561  -1.1984606155      0
#> 562  -0.5101234266      0
#> 563   0.7507622736      0
#> 564  -0.5012426735      0
#> 565   0.0902307786      0
#> 566   0.3273675236      0
#> 567  -0.2974012476      0
#> 568  -2.1345256145      0
#> 569  -0.2039643095      0
#> 570   0.7656878300      0
#> 571  -0.8536118611      0
#> 572  -0.7205155530      0
#> 573  -0.2637254974      0
#> 574  -0.3330516105      0
#> 575   2.2930969454      0
#> 576   0.7066786380      0
#> 577   0.7276627434      0
#> 578  -0.9629581751      0
#> 579   0.8500269927      0
#> 580   0.0035376395      0
#> 581  -0.1278981812      0
#> 582  -0.8696025766      0
#> 583   0.5097315052      0
#> 584   1.3419154841      0
#> 585  -0.7374654700      0
#> 586  -0.7513411563      0
#> 587   0.7503807279      0
#> 588  -0.4500073223      0
#> 589  -0.1569591935      0
#> 590   1.3289814330      0
#> 591  -1.8358037782      0
#> 592   0.1110729905      0
#> 593   0.2359144761      0
#> 594   0.1403149163      0
#> 595  -1.1716554467      0
#> 596  -0.0488585832      0
#> 597  -1.4955839850      0
#> 598   0.6735924884      0
#> 599   0.2365198819      0
#> 600   0.5909498570      0
#> 601  -0.1293618042      0
#> 602  -1.0705125523      0
#> 603   0.0436653643      0
#> 604  -0.9887128609      0
#> 605  -0.6870159218      0
#> 606  -1.0234606434      0
#> 607   1.1439556364      0
#> 608  -0.6976791393      0
#> 609   1.7506712602      0
#> 610   0.7466762717      0
#> 611   0.7416335011      0
#> 612  -1.3711719565      0
#> 613   0.5313453878      0
#> 614  -0.3833041519      0
#> 615   1.6327005920      0
#> 616  -1.1492920048      0
#> 617  -0.8506176740      0
#> 618   1.6591650837      0
#> 619  -0.3737273406      0
#> 620  -1.9570483683      0
#> 621   0.7130889808      0
#> 622   1.3344382113      0
#> 623   2.2359399631      0
#> 624   0.3059152235      0
#> 625  -1.5128801588      0
#> 626   1.5671654854      0
#> 627   1.1331681684      0
#> 628  -1.2761246753      0
#> 629  -0.1183231110      0
#> 630   0.0558617352      0
#> 631  -0.7021790248      0
#> 632  -0.4624180552      0
#> 633  -1.4866173658      0
#> 634   0.9271589419      0
#> 635   1.8514338764      0
#> 636  -0.9759471958      0
#> 637  -1.1710641808      0
#> 638  -1.9053139740      0
#> 639  -0.6084838454      0
#> 640  -0.2921139221      0
#> 641   0.8024937487      0
#> 642  -0.2469210614      0
#> 643  -0.3907444251      0
#> 644   0.0316410732      0
#> 645   0.8560446457      0
#> 646   0.1113062609      0
#> 647  -0.3993064215      0
#> 648  -1.3938708209      0
#> 649  -0.8049729222      0
#> 650  -0.5002726292      0
#> 651  -2.3315051945      0
#> 652   1.1397314113      0
#> 653  -0.9367604046      0
#> 654   1.1927571916      0
#> 655  -0.0448393484      0
#> 656  -2.1175065754      0
#> 657   0.2566050211      0
#> 658  -0.9239052812      0
#> 659  -0.7107924121      0
#> 660  -0.5720110519      0
#> 661  -0.2072218360      0
#> 662   0.7957656203      0
#> 663  -0.3557947126      0
#> 664   1.1931262281      0
#> 665   0.6484782744      0
#> 666   0.7201536536      0
#> 667  -0.6139295195      0
#> 668   0.7360400130      0
#> 669  -1.7107614016      0
#> 670  -0.1080419253      0
#> 671  -0.0909931256      0
#> 672  -1.0396635562      0
#> 673  -0.9031817118      0
#> 674  -0.4990317436      0
#> 675   1.4570330589      0
#> 676   2.5734995845      0
#> 677   1.5207549236      0
#> 678   0.5516490569      0
#> 679   1.5768985927      0
#> 680   0.4863928648      0
#> 681  -0.2408877329      0
#> 682  -1.2230473125      0
#> 683   1.2533771933      0
#> 684   0.1209147523      0
#> 685   0.6441343355      0
#> 686   1.2187353200      0
#> 687   0.5861825809      0
#> 688  -0.3983086962      0
#> 689  -1.0568649635      0
#> 690   0.1680251950      0
#> 691   0.7313616064      0
#> 692   0.7001586709      0
#> 693   1.7101307495      0
#> 694   0.3943952324      0
#> 695  -1.1314593143      0
#> 696  -0.7559094232      0
#> 697  -1.3759520370      0
#> 698  -0.6315126318      0
#> 699   1.4684489054      0
#> 700  -1.0470846168      0
#> 701  -0.1020564060      0
#> 702  -0.4887366127      0
#> 703   2.3265202314      0
#> 704  -0.5195437905      0
#> 705   0.5666831135      0
#> 706   0.2355505129      0
#> 707  -0.1729094424      0
#> 708   0.2690108656      0
#> 709   0.1912319322      0
#> 710   0.0590115473      0
#> 711  -0.3712419967      0
#> 712   0.4015120864      0
#> 713   0.9167536064      0
#> 714  -1.4824209530      0
#> 715   1.1007387653      0
#> 716  -0.4511295471      0
#> 717  -0.3067816812      0
#> 718   0.9499669590      0
#> 719  -0.0519210046      0
#> 720  -0.3019954985      0
#> 721  -0.2486831398      0
#> 722   0.2538024959      0
#> 723  -0.3326738976      0
#> 724  -0.4946136805      0
#> 725  -0.1806980153      0
#> 726   1.3773616458      0
#> 727  -0.8806356695      0
#> 728   1.1613893520      0
#> 729  -1.1653994924      0
#> 730   0.8278700668      0
#> 731  -0.7794121500      0
#> 732   1.3549438173      0
#> 733  -2.2420233572      0
#> 734  -0.6430063966      0
#> 735  -1.9373680576      0
#> 736   0.0405760108      0
#> 737  -0.8653127491      0
#> 738  -0.3481295421      0
#> 739   0.7866846727      0
#> 740   0.4549587140      0
#> 741  -0.8745888696      0
#> 742   0.2171619766      0
#> 743  -0.3422970555      0
#> 744   1.2017942938      0
#> 745   1.2669995258      0
#> 746   0.6224228003      0
#> 747  -0.5036994742      0
#> 748  -2.1121530963      0
#> 749  -0.5401880380      0
#> 750  -1.0168234627      0
#> 751  -0.7703214161      0
#> 752  -0.3539799863      0
#> 753  -1.6974233262      0
#> 754   0.2370950713      0
#> 755  -0.3922006109      0
#> 756  -1.8673086919      0
#> 757   0.4075118103      0
#> 758   0.7421730434      0
#> 759  -0.0252182475      0
#> 760  -1.3609753826      0
#> 761  -1.4674603149      0
#> 762   0.0327398048      0
#> 763  -0.0081622680      0
#> 764  -0.0697609980      0
#> 765   1.3757308334      0
#> 766   1.2526856526      0
#> 767  -1.3412002588      0
#> 768   1.4899128972      0
#> 769   0.9343662702      0
#> 770   0.4599035758      0
#> 771  -0.9220268266      0
#> 772   0.4908888258      0
#> 773  -0.5286231854      0
#> 774   0.1317458237      0
#> 775  -0.1997858930      0
#> 776   0.6633720997      0
#> 777  -1.7691571256      0
#> 778   0.3935291671      0
#> 779   1.6088574857      0
#> 780  -0.3354323800      0
#> 781   1.1734270106      0
#> 782   1.6419579945      0
#> 783   0.2421622736      0
#> 784   0.4419586478      0
#> 785   0.1132142359      0
#> 786   0.9278355110      0
#> 787  -0.3865574257      0
#> 788   0.8495687747      0
#> 789  -1.4751121857      0
#> 790  -0.8402385041      0
#> 791   0.8315802130      0
#> 792  -1.7176953363      0
#> 793  -2.4773252899      0
#> 794   0.5268976607      0
#> 795  -0.4778110708      0
#> 796  -0.6312278626      0
#> 797   0.5024887902      0
#> 798  -0.4904023451      0
#> 799  -0.0190005825      0
#> 800  -0.0713915022      0
#> 801  -0.8428353557      0
#> 802  -0.4348688077      0
#> 803  -2.0237091055      0
#> 804  -0.7093927077      0
#> 805  -0.5908149704      0
#> 806   1.1278138041      0
#> 807   0.6948722683      0
#> 808  -0.5365897437      0
#> 809   1.3586741442      0
#> 810  -0.8880449977      0
#> 811   0.3094101785      0
#> 812  -0.1581177519      0
#> 813  -0.6201317223      0
#> 814   0.6906645196      0
#> 815  -0.1714909020      0
#> 816  -2.7485651002      0
#> 817   1.7430901259      0
#> 818   0.6875797999      0
#> 819  -0.3246110240      0
#> 820  -0.4795962215      0
#> 821   0.4743617979      0
#> 822  -0.1311143635      0
#> 823   1.4066710703      0
#> 824   0.0573760564      0
#> 825  -1.0226341833      0
#> 826   0.1080957349      0
#> 827   0.2690963345      0
#> 828   0.3337708369      0
#> 829  -1.6563338940      0
#> 830   0.5360400591      0
#> 831   0.3769258359      0
#> 832   0.5213489160      0
#> 833   0.7816784227      0
#> 834  -0.6275445700      0
#> 835  -0.8540586470      0
#> 836   0.7862565704      0
#> 837  -0.2816665533      0
#> 838   0.4918734744      0
#> 839   0.1366173751      0
#> 840  -0.5815646400      0
#> 841  -0.1067061582      0
#> 842   0.4705557639      0
#> 843   0.9248736256      0
#> 844  -0.0337259215      0
#> 845  -0.7217312194      0
#> 846   0.9216771926      0
#> 847   1.0184385476      0
#> 848  -1.0325343089      0
#> 849   1.1433729937      0
#> 850   0.6332085414      0
#> 851   0.2769703851      0
#> 852  -2.0957241383      0
#> 853  -0.6406501641      0
#> 854  -0.7548399771      0
#> 855  -1.1700646993      0
#> 856   0.6934697638      0
#> 857   0.6065579677      0
#> 858   0.7943811563      0
#> 859  -1.2470231950      0
#> 860  -1.4750411325      0
#> 861   1.0111229756      0
#> 862   1.2165792551      0
#> 863   1.3452173760      0
#> 864   0.3897204014      0
#> 865  -0.7365610277      0
#> 866  -0.0810274661      0
#> 867   1.5590793664      0
#> 868   0.0938792205      0
#> 869   1.1763912214      0
#> 870   0.5956168747      0
#> 871  -0.6035028180      0
#> 872  -0.4091041407      0
#> 873  -0.8209414079      0
#> 874  -1.1800487507      0
#> 875  -0.4384461357      0
#> 876  -1.5440756292      0
#> 877  -0.1157256816      0
#> 878   0.3785501586      0
#> 879  -0.7529991577      0
#> 880  -0.3983280530      0
#> 881   2.7885954643      0
#> 882   1.3166178999      0
#> 883   2.4111813422      0
#> 884  -0.4214776494      0
#> 885  -1.1052351937      0
#> 886  -0.7321299474      0
#> 887  -0.1112544987      0
#> 888  -1.5401403710      0
#> 889  -0.4715404405      0
#> 890   0.3070260832      0
#> 891   1.6095955547      0
#> 892   0.4503265898      0
#> 893  -1.4721406287      0
#> 894   0.4364987020      0
#> 895   1.2832298652      1
#> 896   0.8046447142      1
#> 897   0.2761067415      1
#> 898   0.1405786502      1
#> 899   1.0942588240      1
#> 900  -0.3023425603      1
#> 901  -0.2513407986      1
#> 902  -0.8238252110      1
#> 903   1.9897065416      1
#> 904   0.3953380030      1
#> 905   0.6350365399      1
#> 906   1.0166864126      1
#> 907   0.0001346672      1
#> 908   1.1506413551      1
#> 909   2.4000346568      1
#> 910  -0.6962851242      1
#> 911   0.4564778514      1
#> 912   0.7688541653      1
#> 913   1.8676630603      1
#> 914  -0.7649013880      1
#> 915  -0.8856158558      1
#> 916   1.4261915833      1
#> 917   0.8694460955      1
#> 918   1.6088501347      1
#> 919   0.6401149492      1
#> 920  -0.3261633521      1
#> 921  -0.0839215640      1
#> 922   1.5067067301      1
#> 923   0.8888145006      1
#> 924  -1.5512106838      1
#> 925  -0.0065952554      1
#> 926   2.8888638113      1
#> 927   0.9557892424      1
#> 928   0.7710283142      1
#> 929   1.0034114296      1
#> 930   0.9656570885      1
#> 931   0.7425133908      1
#> 932   1.6730971272      1
#> 933   0.5098151095      1
#> 934   1.5721788823      1
#> 935   0.2903009324      1
#> 936  -1.6504915204      1
#> 937  -0.2306965048      1
#> 938  -1.2490719551      1
#> 939   1.0237308930      1
#> 940   1.8800245577      1
#> 941   0.4467696543      1
#> 942   2.1600400518      1
#> 943   1.3778207782      1
#> 944   0.3102281591      1
#> 945   1.0232160037      1
#> 946   0.3956292322      1
#> 947   1.5352220732      1
#> 948   0.8603140246      1
#> 949   0.6183833896      1
#> 950   0.6553620527      1
#> 951   0.2666688355      1
#> 952   0.6006246395      1
#> 953   0.7722462978      1
#> 954  -0.1921661534      1
#> 955  -0.4973815489      1
#> 956   2.4405834862      1
#> 957   0.1494094522      1
#> 958   0.9379234873      1
#> 959   3.2328962056      1
#> 960   0.2167404594      1
#> 961   1.0428955240      1
#> 962  -0.4284775820      1
#> 963   1.4113560574      1
#> 964   0.6867186990      1
#> 965   1.4400149930      1
#> 966   2.9511401941      1
#> 967   0.8060364770      1
#> 968   0.4723720406      1
#> 969   1.4983341944      1
#> 970  -0.8369003462      1
#> 971   1.1125062078      1
#> 972  -0.0975664409      1
#> 973   1.0047947225      1
#> 974   0.8465263328      1
#> 975  -0.7464633544      1
#> 976   1.9724569930      1
#> 977   0.7766028209      1
#> 978   3.1332446834      1
#> 979  -1.1505886672      1
#> 980   2.6076410337      1
#> 981   1.4150044139      1
#> 982   1.7321309111      1
#> 983   0.5677871615      1
#> 984  -0.5374404642      1
#> 985   1.6023744680      1
#> 986  -1.6129202901      1
#> 987  -1.0388786998      1
#> 988   0.6205573416      1
#> 989   0.2467678781      1
#> 990   0.5901053194      1
#> 991   0.8460278515      1
#> 992   0.3542864579      1
#> 993  -0.0808893501      1
#> 994   0.5737825552      1
#> 995  -0.1908900846      1
#> 996  -0.0584207118      1
#> 997   1.1945154741      1
#> 998   1.1180320930      1
#> 999  -0.2574938943      1
#> 1000  0.4772605237      1