Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     0.6895768387      0
#> 2    -0.7831492029      0
#> 3    -0.2297780099      0
#> 4     0.5233467751      0
#> 5    -0.2670320640      0
#> 6     0.4733724620      0
#> 7     0.0312785845      0
#> 8    -0.2847832010      0
#> 9    -0.7562322636      0
#> 10    0.1661493560      0
#> 11    2.4278149777      0
#> 12    0.6434696955      0
#> 13   -0.5466112230      0
#> 14   -0.5557508489      0
#> 15   -0.5999181884      0
#> 16   -0.7648732370      0
#> 17    0.2488835457      0
#> 18    0.8383721616      0
#> 19    2.0318535079      0
#> 20    0.4779884696      0
#> 21   -1.3562977306      0
#> 22   -0.2506484137      0
#> 23    0.2714762256      0
#> 24    0.3616299089      0
#> 25   -0.2814683168      0
#> 26   -1.2147579855      0
#> 27    0.1411724480      0
#> 28   -1.1787936826      0
#> 29    2.2037944778      0
#> 30    0.8819118975      0
#> 31   -1.0417105756      0
#> 32   -0.6370074006      0
#> 33    0.8957104817      0
#> 34    0.8978499534      0
#> 35   -0.7240068050      0
#> 36    0.1790643658      0
#> 37    0.2586568720      0
#> 38    3.0511241302      0
#> 39    0.8703951414      0
#> 40   -0.5395893944      0
#> 41    0.3895565841      0
#> 42   -0.2844825020      0
#> 43   -0.2208772089      0
#> 44   -0.5073878977      0
#> 45    1.3349815024      0
#> 46   -0.4996336195      0
#> 47   -0.9414654420      0
#> 48    0.8813417849      0
#> 49   -0.4320993254      0
#> 50   -0.3048606027      0
#> 51   -0.6097085959      0
#> 52    0.0896736898      0
#> 53   -0.8904092503      0
#> 54    0.1087780493      0
#> 55    1.2270601840      0
#> 56   -0.6228114942      0
#> 57    0.4768391660      0
#> 58   -0.3519246104      0
#> 59   -1.6657093104      0
#> 60    1.5338982051      0
#> 61    1.4247997792      0
#> 62    0.6647340525      0
#> 63    1.8814376053      0
#> 64   -0.5140664172      0
#> 65   -1.1792646004      0
#> 66    0.3777141316      0
#> 67   -0.5846554217      0
#> 68    0.4506385323      0
#> 69   -1.6656045528      0
#> 70    0.2721705198      0
#> 71    0.4526550407      0
#> 72   -0.3634845744      0
#> 73    1.5305954829      0
#> 74    0.5140617711      0
#> 75   -1.7647683075      0
#> 76   -0.2112962767      0
#> 77   -0.9009940335      0
#> 78    0.1984029070      0
#> 79    0.4030891903      0
#> 80    0.0841874889      0
#> 81   -1.9007749403      0
#> 82    0.2159941195      0
#> 83    0.0513720762      0
#> 84    0.5328731821      0
#> 85    3.2737993783      0
#> 86   -1.5105340273      0
#> 87    0.4703381484      0
#> 88    1.3811289594      0
#> 89    1.0649654541      0
#> 90    0.2167354626      0
#> 91    0.4363412406      0
#> 92   -1.0276857937      0
#> 93   -1.1004968306      0
#> 94    1.7603502465      0
#> 95   -0.3099764842      0
#> 96    0.5780251899      0
#> 97   -0.2899485509      0
#> 98    1.2533619340      0
#> 99   -0.1400330049      0
#> 100   0.2748580412      0
#> 101  -0.1315004118      0
#> 102   0.2584336144      0
#> 103   0.7502191308      0
#> 104  -0.0194917209      0
#> 105   0.4383546011      0
#> 106   0.1610862564      0
#> 107  -0.8264925318      0
#> 108   2.1510432987      0
#> 109   0.5974902044      0
#> 110   1.2542963856      0
#> 111  -0.7080145073      0
#> 112  -0.5916705509      0
#> 113  -2.0807148830      0
#> 114   0.1695303278      0
#> 115   0.1432167316      0
#> 116  -0.6066376929      0
#> 117   0.7798269822      0
#> 118   0.6361978377      0
#> 119  -0.2600973286      0
#> 120  -1.6042888451      0
#> 121  -0.3292437784      0
#> 122   0.5008786853      0
#> 123  -1.5665420775      0
#> 124   0.4076505139      0
#> 125  -0.8110063132      0
#> 126  -0.0780216325      0
#> 127   0.5597521844      0
#> 128   0.0732864415      0
#> 129  -0.9939344188      0
#> 130  -0.1358894315      0
#> 131   0.1852518665      0
#> 132  -1.9249012236      0
#> 133  -1.2779670861      0
#> 134   0.1484435812      0
#> 135  -0.9836147147      0
#> 136  -1.0482624224      0
#> 137   0.1529403379      0
#> 138   1.0818611661      0
#> 139  -1.0726429313      0
#> 140   1.5240222974      0
#> 141   0.7109792398      0
#> 142  -1.1138385176      0
#> 143  -0.2538624123      0
#> 144  -1.0776117255      0
#> 145  -0.3245616736      0
#> 146   0.4195874156      0
#> 147  -0.8675091973      0
#> 148   0.4266097460      0
#> 149  -1.1434826020      0
#> 150  -0.4809467473      0
#> 151   0.7829344289      0
#> 152  -0.2610548617      0
#> 153  -0.7447785552      0
#> 154  -0.7936267600      0
#> 155  -1.9747990027      0
#> 156  -0.8454120657      0
#> 157   0.2415869342      0
#> 158  -0.0164582493      0
#> 159   0.6177448761      0
#> 160   0.1638785636      0
#> 161   0.6381743398      0
#> 162   0.0827717190      0
#> 163   0.5643057382      0
#> 164   0.5454413996      0
#> 165   0.8451402318      0
#> 166  -0.1779374979      0
#> 167   1.8176295294      0
#> 168   0.8429892537      0
#> 169  -0.3785341041      0
#> 170  -0.1978508691      0
#> 171  -0.6487125334      0
#> 172   0.4205461040      0
#> 173   0.6007613095      0
#> 174   0.2224735583      0
#> 175  -0.2897758982      0
#> 176  -1.0964682199      0
#> 177  -0.1806245009      0
#> 178   0.1418579932      0
#> 179   0.6255087828      0
#> 180  -0.5007961551      0
#> 181  -0.0684920299      0
#> 182   0.3398982971      0
#> 183   1.4522687822      0
#> 184  -0.1897624951      0
#> 185   0.3068094833      0
#> 186  -0.2219698849      0
#> 187  -0.9372060928      0
#> 188   0.2690924497      0
#> 189   0.5619619012      0
#> 190  -1.1541989130      0
#> 191  -0.0962354926      0
#> 192   0.6510658543      0
#> 193   0.5985747843      0
#> 194  -1.2654765309      0
#> 195  -0.0792659468      0
#> 196  -0.4040475487      0
#> 197   0.2334467769      0
#> 198  -0.7282223840      0
#> 199  -0.0826448180      0
#> 200  -0.3711927333      0
#> 201  -0.8716847270      0
#> 202   1.3085781541      0
#> 203   1.3336026582      0
#> 204  -0.1024897975      0
#> 205   0.3094424126      0
#> 206   0.6370053339      0
#> 207  -1.2925573644      0
#> 208  -1.8170595607      0
#> 209  -1.4243320220      0
#> 210   1.6552605836      0
#> 211   0.4060048088      0
#> 212  -0.7626827147      0
#> 213  -0.6040434954      0
#> 214   0.0476657139      0
#> 215   2.1941909492      0
#> 216   1.9784380054      0
#> 217  -0.7844098215      0
#> 218  -0.8155683845      0
#> 219  -0.0144969409      0
#> 220  -0.9464513274      0
#> 221  -1.4281228914      0
#> 222  -0.3238885640      0
#> 223  -0.4611036561      0
#> 224   0.1593065758      0
#> 225  -0.5580106095      0
#> 226   1.3968354804      0
#> 227  -1.2578651848      0
#> 228   0.0837540852      0
#> 229   0.3364271420      0
#> 230   1.6290958101      0
#> 231   0.6456455970      0
#> 232  -1.5423217905      0
#> 233   1.8184769416      0
#> 234  -0.2589553036      0
#> 235  -0.5015577710      0
#> 236   1.4708000504      0
#> 237   0.5278633179      0
#> 238  -1.2912656890      0
#> 239   1.0474753393      0
#> 240  -0.9551703863      0
#> 241   0.2525012881      0
#> 242   0.0500906689      0
#> 243  -0.4628567853      0
#> 244  -0.9853528971      0
#> 245   0.6484308403      0
#> 246  -0.4002090135      0
#> 247  -0.1647527776      0
#> 248  -1.1763058891      0
#> 249   1.4921598584      0
#> 250   1.7846546970      0
#> 251   0.3788893332      0
#> 252   1.5854454415      0
#> 253  -0.4941579831      0
#> 254   1.0086073926      0
#> 255   0.5297245486      0
#> 256   0.4219602890      0
#> 257  -0.3292318062      0
#> 258  -1.2424678026      0
#> 259   1.0037975320      0
#> 260   1.5594904407      0
#> 261  -0.3258281375      0
#> 262  -0.5601431827      0
#> 263  -0.4046641875      0
#> 264   0.3781091272      0
#> 265   1.2295351306      0
#> 266   1.1108080027      0
#> 267  -0.3555667542      0
#> 268  -0.8598482196      0
#> 269  -0.5402720295      0
#> 270   1.0277777662      0
#> 271   0.7407699122      0
#> 272  -0.9799765096      0
#> 273   0.4114060302      0
#> 274   2.0483891652      0
#> 275  -0.5438531276      0
#> 276   0.3517184135      0
#> 277  -1.4958539736      0
#> 278   0.5201342057      0
#> 279   1.5499206363      0
#> 280   0.5065033364      0
#> 281   0.9245687368      0
#> 282   0.3938373941      0
#> 283  -0.2818865904      0
#> 284  -1.1034670894      0
#> 285  -0.7146286058      0
#> 286  -0.4042098684      0
#> 287   1.0683598819      0
#> 288   0.7036926225      0
#> 289   1.4647042176      0
#> 290   0.5120223954      0
#> 291  -0.3719906731      0
#> 292   0.1828255248      0
#> 293  -0.7807066076      0
#> 294   1.8758395149      0
#> 295  -1.2231316339      0
#> 296  -1.4075052518      0
#> 297  -0.1399537108      0
#> 298   0.1086284201      0
#> 299   0.3114166142      0
#> 300  -0.9115802133      0
#> 301   1.4928126133      0
#> 302   0.7190338379      0
#> 303   0.7129966733      0
#> 304  -0.6002804944      0
#> 305   0.6199003022      0
#> 306  -0.0860181181      0
#> 307   0.2626729517      0
#> 308  -0.2879344970      0
#> 309  -0.9551785565      0
#> 310  -0.7341044354      0
#> 311   1.0165589982      0
#> 312  -2.4033198801      0
#> 313  -1.3967600558      0
#> 314  -1.2158122519      0
#> 315  -0.0891569539      0
#> 316   0.5776282479      0
#> 317   1.1023220959      0
#> 318  -1.0399915275      0
#> 319  -1.6015469126      0
#> 320  -0.1952614719      0
#> 321   1.5298102865      0
#> 322   0.1505858399      0
#> 323   1.5069275979      0
#> 324  -0.0450799553      0
#> 325  -0.8810352358      0
#> 326   0.2542471593      0
#> 327  -1.5752646435      0
#> 328   0.7004268092      0
#> 329   0.0560889448      0
#> 330  -0.0188290304      0
#> 331  -0.0388635498      0
#> 332  -2.3373146033      0
#> 333  -1.9453637623      0
#> 334   0.1413120923      0
#> 335   1.3168249235      0
#> 336  -0.0475528238      0
#> 337   1.0979671842      0
#> 338   1.8610173321      0
#> 339   1.8253512897      0
#> 340   0.7954378869      0
#> 341  -0.5744509660      0
#> 342  -0.6914452017      0
#> 343   0.0661700942      0
#> 344   0.8807633741      0
#> 345   0.7401452035      0
#> 346  -1.7173777215      0
#> 347  -0.9933575565      0
#> 348  -0.1879669791      0
#> 349  -0.4905337901      0
#> 350   0.4881766547      0
#> 351  -1.5670651433      0
#> 352  -1.5492100103      0
#> 353  -0.3390877185      0
#> 354   0.3024911988      0
#> 355   0.4373610072      0
#> 356  -0.2415570356      0
#> 357  -0.9807727584      0
#> 358  -0.6672127805      0
#> 359  -0.3169721694      0
#> 360   1.7515729968      0
#> 361   0.4653136914      0
#> 362   1.1394611732      0
#> 363  -0.1462527291      0
#> 364   1.6017456414      0
#> 365   1.6927968997      0
#> 366  -0.1135451422      0
#> 367  -0.8220701139      0
#> 368   0.0875375418      0
#> 369  -1.5906695981      0
#> 370   1.6286555723      0
#> 371  -0.8318240565      0
#> 372  -0.6811338293      0
#> 373   0.6334523752      0
#> 374  -0.2891790037      0
#> 375  -0.6503207060      0
#> 376  -1.5620761922      0
#> 377   1.0587453202      0
#> 378   1.4946149660      0
#> 379  -1.2404673167      0
#> 380  -1.0962585861      0
#> 381  -0.7398947805      0
#> 382  -0.5155195089      0
#> 383  -1.1363651528      0
#> 384   0.3954389394      0
#> 385  -0.3868576454      0
#> 386   0.8910321509      0
#> 387   0.6436755153      0
#> 388  -0.8390406182      0
#> 389   0.6904603115      0
#> 390   0.4322021591      0
#> 391   1.3507265228      0
#> 392  -0.6791980008      0
#> 393   0.4464272309      0
#> 394   0.2309149147      0
#> 395  -0.0939616972      0
#> 396  -1.8410014146      0
#> 397   1.6763206819      0
#> 398   0.4180235301      0
#> 399  -0.6761830895      0
#> 400   0.8409779155      0
#> 401   1.7621777452      0
#> 402   0.2217498462      0
#> 403  -1.7432192754      0
#> 404  -0.1345540473      0
#> 405   0.5764816012      0
#> 406   0.7665028386      0
#> 407   0.3815948845      0
#> 408  -0.6910275454      0
#> 409   0.2064782377      0
#> 410  -1.1502403506      0
#> 411   0.5907346674      0
#> 412   1.1201061020      0
#> 413  -0.8926594203      0
#> 414  -0.0573117481      0
#> 415  -0.2560129721      0
#> 416   0.0122815446      0
#> 417  -0.9212084742      0
#> 418   1.1936275613      0
#> 419   0.4537842777      0
#> 420   1.0592547144      0
#> 421  -0.0619927091      0
#> 422   0.7467585913      0
#> 423  -0.0258810589      0
#> 424  -0.0928033283      0
#> 425   0.7219628899      0
#> 426  -0.4315243401      0
#> 427  -0.0204377751      0
#> 428   1.7794166911      0
#> 429   0.9128563157      0
#> 430   1.5287099406      0
#> 431   1.2606360394      0
#> 432   0.8013893696      0
#> 433   1.0245651231      0
#> 434   0.4465692657      0
#> 435  -1.2547692000      0
#> 436  -1.0368258545      0
#> 437   1.8768749309      0
#> 438   0.6094809255      0
#> 439   1.5874199750      0
#> 440  -1.0720143071      0
#> 441  -1.0019745394      0
#> 442   0.9445460798      0
#> 443  -0.4604938661      0
#> 444   1.0860701151      0
#> 445  -0.7328998359      0
#> 446   0.6214290759      0
#> 447   0.5102119241      0
#> 448   0.6605994595      0
#> 449  -1.5304700914      0
#> 450   0.1349349899      0
#> 451   0.2228533703      0
#> 452   0.7932651443      0
#> 453   0.1600063110      0
#> 454   0.5058722355      0
#> 455  -0.4991741633      0
#> 456  -0.0172949016      0
#> 457  -0.1650228120      0
#> 458  -1.0093111407      0
#> 459   0.3455833453      0
#> 460  -0.3963675124      0
#> 461  -0.6949896240      0
#> 462  -1.6832219462      0
#> 463   1.3494029054      0
#> 464   0.6076772616      0
#> 465   0.2304572481      0
#> 466  -0.6668547384      0
#> 467  -0.9431274048      0
#> 468  -0.6719219656      0
#> 469   0.6996069474      0
#> 470  -0.7667329885      0
#> 471  -1.2267707248      0
#> 472   2.4535103855      0
#> 473  -0.1905413038      0
#> 474   0.5586238435      0
#> 475  -0.3704739872      0
#> 476   0.3315665674      0
#> 477   0.8974394799      0
#> 478   1.7911476556      0
#> 479   0.8932808807      0
#> 480  -0.0129675826      0
#> 481  -0.0867547779      0
#> 482   1.1042661052      0
#> 483  -0.1249535159      0
#> 484  -0.0180341574      0
#> 485   0.0684669936      0
#> 486  -0.0391629166      0
#> 487   1.0705068310      0
#> 488   0.2779738381      0
#> 489   0.5869162830      0
#> 490   2.0479959932      0
#> 491  -1.6070832001      0
#> 492   0.1933602619      0
#> 493   1.1494210880      0
#> 494   1.7074240431      0
#> 495   0.1474330073      0
#> 496   1.0091800103      0
#> 497  -0.1458704100      0
#> 498  -0.6710071199      0
#> 499   0.7069844121      0
#> 500  -0.1559105736      0
#> 501   0.9060829786      0
#> 502  -1.2243727648      0
#> 503   0.0776994058      0
#> 504   0.5482397901      0
#> 505   0.0035544629      0
#> 506   0.5869598580      0
#> 507   1.1448425161      0
#> 508   0.1778347553      0
#> 509  -1.2500804442      0
#> 510  -0.1383311946      0
#> 511   0.7669843508      0
#> 512  -0.4271262853      0
#> 513   1.1719770687      0
#> 514  -0.9521964934      0
#> 515   0.6535089764      0
#> 516  -0.2025216276      0
#> 517   0.2296492065      0
#> 518   0.5156997127      0
#> 519   1.1981649049      0
#> 520   1.7992313838      0
#> 521   0.2864162164      0
#> 522  -1.1395271867      0
#> 523   0.5415698161      0
#> 524   0.2886034611      0
#> 525   0.2235557989      0
#> 526   0.9363594246      0
#> 527   0.5277752937      0
#> 528   0.0745780532      0
#> 529   0.5987292104      0
#> 530  -1.5750386158      0
#> 531   0.2767983191      0
#> 532   0.5574643448      0
#> 533  -0.8393754784      0
#> 534  -2.0610669062      0
#> 535   1.0281332422      0
#> 536  -0.3698810897      0
#> 537   2.1847677651      0
#> 538   0.1667763181      0
#> 539   0.4879156189      0
#> 540   0.7467358922      0
#> 541   1.5998636637      0
#> 542   0.5315750841      0
#> 543  -2.0329650575      0
#> 544  -0.9971735105      0
#> 545  -0.4017001970      0
#> 546   0.2658714361      0
#> 547   2.7783380866      0
#> 548  -0.0564642451      0
#> 549   0.1639287006      0
#> 550   0.6887018371      0
#> 551   1.0941039111      0
#> 552  -1.3001938221      0
#> 553  -0.1329454711      0
#> 554  -0.8217179194      0
#> 555   1.3301173679      0
#> 556   1.6230998311      0
#> 557   0.5419780663      0
#> 558  -0.5655909412      0
#> 559   0.1641838428      0
#> 560   0.6255621975      0
#> 561   0.9486593141      0
#> 562   1.8756606258      0
#> 563  -0.6260037758      0
#> 564  -1.8810711478      0
#> 565   0.1252442371      0
#> 566  -1.0440584194      0
#> 567   0.4480697537      0
#> 568  -0.8045432093      0
#> 569  -0.3768228152      0
#> 570   0.5864513309      0
#> 571  -1.3273587769      0
#> 572  -1.6765975335      0
#> 573  -0.5521251438      0
#> 574   0.2362815362      0
#> 575   0.5041346892      0
#> 576   1.2780845325      0
#> 577  -0.1766705087      0
#> 578  -2.7213592304      0
#> 579  -0.8056168481      0
#> 580  -0.2599731624      0
#> 581   1.0424144169      0
#> 582  -0.1513902250      0
#> 583   0.6337045348      0
#> 584  -1.3010504675      0
#> 585   0.9794586942      0
#> 586   0.1214509323      0
#> 587   0.9182245428      0
#> 588   0.5339273644      0
#> 589  -0.4610360646      0
#> 590   1.1821271334      0
#> 591   0.9557480081      0
#> 592  -0.0467444748      0
#> 593  -1.1785930018      0
#> 594  -0.9794929037      0
#> 595  -0.9204386108      0
#> 596  -0.8081033938      0
#> 597  -0.9785057318      0
#> 598   1.7425974464      0
#> 599   1.0611797508      0
#> 600   0.4995406003      0
#> 601   1.8813220841      0
#> 602   1.4545879958      0
#> 603  -0.5671103871      0
#> 604  -1.3782286604      0
#> 605  -0.7898987643      0
#> 606   0.4698840592      0
#> 607  -1.1125206619      0
#> 608  -0.3091975661      0
#> 609  -2.0039039331      0
#> 610  -0.7207460132      0
#> 611   0.4808869782      0
#> 612   2.1484264415      0
#> 613  -0.5710765791      0
#> 614  -1.3469981658      0
#> 615  -1.3164855137      0
#> 616  -1.2323108283      0
#> 617   0.7868410171      0
#> 618  -1.6673499088      0
#> 619  -1.6365618288      0
#> 620  -0.8379428672      0
#> 621  -0.1061620323      0
#> 622  -1.9781420071      0
#> 623  -1.0683643013      0
#> 624  -0.8506227976      0
#> 625   0.3212112679      0
#> 626   0.3994357323      0
#> 627   0.7557246989      0
#> 628   0.9123707431      0
#> 629   0.4976363969      0
#> 630  -0.1905669984      0
#> 631  -0.0485358261      0
#> 632  -1.5318828105      0
#> 633   1.0164103225      0
#> 634  -0.3353108708      0
#> 635   1.5464229317      0
#> 636   0.4457298961      0
#> 637  -0.9078882263      0
#> 638  -1.1995418326      0
#> 639   0.0617787533      0
#> 640  -0.6201499700      0
#> 641  -0.3288978339      0
#> 642  -0.0799299443      0
#> 643  -0.6733398669      0
#> 644  -1.5291226247      0
#> 645   1.3835334189      0
#> 646   1.0697050031      0
#> 647   0.3705877158      0
#> 648  -0.4127589986      0
#> 649  -0.2377209580      0
#> 650   0.6570851498      0
#> 651   1.7752745704      0
#> 652  -0.4994306691      0
#> 653   0.5805800873      0
#> 654   0.2481780681      0
#> 655  -1.0215409847      0
#> 656   0.7140918602      0
#> 657   1.1971576745      0
#> 658   0.0372178588      0
#> 659  -0.6584909336      0
#> 660   0.9325084954      0
#> 661   0.8008382797      0
#> 662  -0.5196177188      0
#> 663  -0.6981007908      0
#> 664   1.1579053544      0
#> 665  -0.9311543575      0
#> 666  -1.3156571006      0
#> 667  -0.2529515875      0
#> 668  -0.7193851026      0
#> 669   0.0338060217      0
#> 670   0.1652853566      0
#> 671  -0.0249729506      0
#> 672   0.1834459417      0
#> 673   0.4024961756      0
#> 674  -0.0405719392      0
#> 675  -1.6342461113      0
#> 676  -0.0855260571      0
#> 677   0.3553275678      0
#> 678   0.1800887365      0
#> 679   0.1104971148      0
#> 680   0.4577539585      0
#> 681   0.1983651826      0
#> 682  -0.4401782256      0
#> 683  -0.1365054885      0
#> 684  -0.0661281667      0
#> 685  -0.3657482905      0
#> 686  -0.3814373173      0
#> 687  -1.4605970814      0
#> 688  -0.3144861485      0
#> 689   0.3744464421      0
#> 690  -1.3381589169      0
#> 691  -1.4293897039      0
#> 692   0.8363222691      0
#> 693   0.8520256451      0
#> 694  -1.3604388869      0
#> 695   1.6943061888      0
#> 696  -0.6216867246      0
#> 697  -0.5067217363      0
#> 698  -0.7446111253      0
#> 699   0.0386371402      0
#> 700   0.9799747539      0
#> 701   0.3020296680      0
#> 702   2.3896529428      0
#> 703   0.9921295375      0
#> 704  -0.7580689473      0
#> 705   1.3958750075      0
#> 706   1.0425376728      0
#> 707   0.0333273960      0
#> 708   0.4020658844      0
#> 709   0.6700504087      0
#> 710  -1.6788225807      0
#> 711  -1.0837800952      0
#> 712  -0.0094435410      0
#> 713  -0.9888200048      0
#> 714   0.4993108193      0
#> 715   0.7310996109      0
#> 716   0.8198416185      0
#> 717  -2.0035574659      0
#> 718   0.8225157325      0
#> 719   1.0352313198      0
#> 720  -0.5584590635      0
#> 721  -0.0998503090      0
#> 722  -0.8457585635      0
#> 723  -1.1593598295      0
#> 724  -0.1854786892      0
#> 725  -0.1115600425      0
#> 726  -1.1027306036      0
#> 727  -0.7594114491      0
#> 728  -0.6459143905      0
#> 729  -0.4309273913      0
#> 730   0.3978375646      0
#> 731  -0.5838793016      0
#> 732  -1.8441575091      0
#> 733  -1.0572327457      0
#> 734   0.5864101603      0
#> 735  -0.9934533061      0
#> 736   0.2566263083      0
#> 737  -0.4623077314      0
#> 738  -0.2311266796      0
#> 739  -0.6677932398      0
#> 740   1.5246369351      0
#> 741  -0.1250803332      0
#> 742  -0.5959897653      0
#> 743  -0.7650706733      0
#> 744   1.7170169663      0
#> 745   0.1806031551      0
#> 746  -1.6545051012      0
#> 747  -1.3184968529      0
#> 748   0.3348658279      0
#> 749  -0.3917928488      0
#> 750  -0.5890394016      0
#> 751  -0.5668545310      0
#> 752   0.6172989141      0
#> 753  -1.9842326649      0
#> 754  -0.6005232871      0
#> 755  -0.6837282278      0
#> 756  -0.9734158997      0
#> 757   0.2152368056      0
#> 758  -0.9118782789      0
#> 759  -0.2127067601      0
#> 760   0.1538467707      0
#> 761  -1.0451268662      0
#> 762   0.3246198219      0
#> 763   0.0670882756      0
#> 764  -0.8379111768      0
#> 765   0.9824691315      0
#> 766   0.5226094461      0
#> 767  -0.9784816083      0
#> 768  -0.5482905975      0
#> 769   0.8582473789      0
#> 770   0.2449392732      0
#> 771  -0.9005304671      0
#> 772  -0.1522446981      0
#> 773  -0.8561608146      0
#> 774  -1.3120866450      0
#> 775  -0.6026782836      0
#> 776   0.2970541786      0
#> 777  -1.2083518398      0
#> 778  -0.0323546892      0
#> 779  -0.5505495032      0
#> 780  -0.4146913145      0
#> 781  -0.4895810873      0
#> 782   0.3657915772      0
#> 783  -0.2202698868      0
#> 784  -0.3062658227      0
#> 785   0.5797589886      0
#> 786   0.1311187054      0
#> 787  -1.5831558266      0
#> 788  -0.6113643091      0
#> 789  -0.8680689946      0
#> 790  -0.0899328305      0
#> 791   3.2932293135      0
#> 792  -0.8470320519      0
#> 793   1.5757947375      0
#> 794   0.0288072379      0
#> 795   0.0621555447      0
#> 796  -0.3103939075      0
#> 797   0.0557184282      0
#> 798   0.7656444283      0
#> 799  -0.6477213673      0
#> 800  -0.9711956522      0
#> 801  -1.5183099644      0
#> 802  -1.3330396534      0
#> 803  -0.2167259479      0
#> 804   1.3686254608      0
#> 805   0.1300515599      0
#> 806  -0.2925636633      0
#> 807  -0.4856207615      0
#> 808  -0.3252331964      0
#> 809   0.5906347703      0
#> 810  -1.1313201000      0
#> 811  -0.1060992391      0
#> 812   0.0343618783      0
#> 813  -0.2113246567      0
#> 814   0.4610695632      0
#> 815   0.0242376899      0
#> 816  -0.0313718541      0
#> 817   0.8250361444      0
#> 818  -0.5511049389      0
#> 819   0.3563881440      0
#> 820   1.9651516494      0
#> 821   1.2492345785      0
#> 822   0.7805095716      0
#> 823   0.7773871653      0
#> 824  -0.4753908646      0
#> 825   0.3917459026      0
#> 826  -0.0285530411      0
#> 827   0.9839619282      0
#> 828  -1.4914924607      0
#> 829   1.5700289860      0
#> 830  -0.4793770604      0
#> 831   1.4095760335      0
#> 832   1.2562076692      0
#> 833   0.4816271109      0
#> 834   0.2374217353      0
#> 835   0.7370803967      0
#> 836   1.0213212531      0
#> 837   0.8244375818      0
#> 838   0.0652025016      0
#> 839  -0.1450735307      0
#> 840   1.4456227552      0
#> 841  -0.8463122635      0
#> 842   0.5066723922      0
#> 843   0.0286569007      0
#> 844  -0.8804353498      0
#> 845  -1.8058927924      0
#> 846  -1.4669323158      0
#> 847   1.1209985495      0
#> 848   0.0085914381      0
#> 849   0.7728075880      0
#> 850   0.7671737656      0
#> 851   0.2009723190      0
#> 852  -0.7073080506      0
#> 853   1.4265250591      0
#> 854   1.2157460098      0
#> 855   2.2072093889      0
#> 856   0.6208407106      0
#> 857  -0.2732731336      0
#> 858  -0.6519073084      0
#> 859   0.9768327980      0
#> 860   0.5197976359      0
#> 861   0.7761951015      0
#> 862  -0.6745221029      0
#> 863   0.4932488644      0
#> 864   1.5021093161      0
#> 865  -0.8419956019      0
#> 866   0.7568835627      0
#> 867  -0.9068379342      0
#> 868   0.7346324272      0
#> 869   0.1108448798      0
#> 870  -0.6314652983      0
#> 871  -0.3242818930      0
#> 872   0.0155071780      0
#> 873   1.8562904746      0
#> 874   0.3694213902      0
#> 875   0.1682067319      0
#> 876  -2.1851333760      0
#> 877   1.0242799372      0
#> 878  -0.5107864947      0
#> 879   0.0614268552      0
#> 880  -1.0919732156      0
#> 881   0.1760493210      0
#> 882  -0.3278093202      0
#> 883   0.0706219719      0
#> 884  -0.7552662949      0
#> 885  -0.3108238551      0
#> 886  -0.1184026921      0
#> 887   0.9820769272      0
#> 888  -1.4290975896      0
#> 889   1.1152438521      0
#> 890   0.5032992610      0
#> 891   1.2450913140      0
#> 892   0.6768762232      0
#> 893  -0.7780697867      0
#> 894   0.2692525932      0
#> 895  -0.3253642498      0
#> 896  -2.8817012454      0
#> 897   0.0006617947      0
#> 898   1.2645908780      0
#> 899   2.0176877498      0
#> 900   2.1541844094      0
#> 901  -0.7582112679      0
#> 902   0.1190293823      0
#> 903   0.8052896210      0
#> 904   0.1827051763      0
#> 905   0.0413001595      0
#> 906  -0.8128856961      0
#> 907   1.9230204869      1
#> 908   1.2381071778      1
#> 909   3.2143429800      1
#> 910   0.1383466751      1
#> 911   0.7432443072      1
#> 912   0.5080292022      1
#> 913   2.2118092515      1
#> 914   0.1755925338      1
#> 915   0.3664628500      1
#> 916   0.8918129232      1
#> 917   1.8461890317      1
#> 918   1.6988573148      1
#> 919  -0.6074619716      1
#> 920   2.5793450966      1
#> 921   0.9638464108      1
#> 922   0.7948208196      1
#> 923   1.6617106794      1
#> 924   1.2680222029      1
#> 925  -0.0187576613      1
#> 926  -0.6138126221      1
#> 927   0.3883490859      1
#> 928   0.4811676418      1
#> 929   0.2807533381      1
#> 930  -0.0259149392      1
#> 931   2.1919505205      1
#> 932   1.5692220939      1
#> 933   0.8873638450      1
#> 934   1.1014824609      1
#> 935  -0.2751014683      1
#> 936   1.5009373160      1
#> 937   0.1324234569      1
#> 938  -0.3574572597      1
#> 939   1.0717735898      1
#> 940   1.0672717723      1
#> 941  -0.2589529818      1
#> 942   0.9274221370      1
#> 943   1.6700478249      1
#> 944   0.5206164821      1
#> 945   0.1103011933      1
#> 946   0.0660885236      1
#> 947   1.0714991703      1
#> 948   2.5112291053      1
#> 949   0.5990420677      1
#> 950   1.6916377923      1
#> 951   1.1417908418      1
#> 952   0.4917269469      1
#> 953   0.1718898755      1
#> 954   0.1684359414      1
#> 955   1.5979800216      1
#> 956   1.7365013760      1
#> 957   1.0722869731      1
#> 958   0.8145676503      1
#> 959  -1.7633555390      1
#> 960   1.3359359337      1
#> 961   0.7114563999      1
#> 962   2.1069291390      1
#> 963   2.3100966677      1
#> 964   0.4480661117      1
#> 965   0.8285840347      1
#> 966   0.9854465311      1
#> 967   0.2897200518      1
#> 968   1.2835111105      1
#> 969   0.4038195782      1
#> 970   0.4359608002      1
#> 971  -0.3764192141      1
#> 972  -0.5674150570      1
#> 973   0.9434773689      1
#> 974  -0.0639673184      1
#> 975   0.7556771099      1
#> 976   1.2159991617      1
#> 977  -1.2002042960      1
#> 978   2.2716272941      1
#> 979   1.1354850761      1
#> 980   1.0682688486      1
#> 981   1.2977327289      1
#> 982   1.2048879776      1
#> 983   0.8074683647      1
#> 984  -0.8607985277      1
#> 985   0.8717851964      1
#> 986   0.3731929579      1
#> 987   1.0624939137      1
#> 988   0.5849212505      1
#> 989   0.2974042274      1
#> 990   1.3417332971      1
#> 991   1.4020540684      1
#> 992   0.8126799004      1
#> 993   0.3500072530      1
#> 994   1.0928644998      1
#> 995   0.6060739120      1
#> 996   1.6624664393      1
#> 997   0.3821443095      1
#> 998   3.5417860419      1
#> 999   1.0692121544      1
#> 1000  0.4989192416      1