Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.288966732      0
#> 2    -0.589592102      0
#> 3     0.799008927      0
#> 4    -1.765369742      0
#> 5    -2.169692750      0
#> 6    -0.759329898      0
#> 7     2.410873853      0
#> 8    -0.460020025      0
#> 9    -1.524220950      0
#> 10    1.228735814      0
#> 11   -0.716878245      0
#> 12    0.702164878      0
#> 13   -2.122659805      0
#> 14   -0.031138240      0
#> 15   -0.044818791      0
#> 16    0.774934922      0
#> 17    0.155250329      0
#> 18    1.096064909      0
#> 19   -1.820497277      0
#> 20    0.128623559      0
#> 21   -1.226811637      0
#> 22    1.309714838      0
#> 23    1.158561179      0
#> 24   -0.363668923      0
#> 25   -0.136961088      0
#> 26    0.110325258      0
#> 27   -0.221523095      0
#> 28   -2.076810566      0
#> 29    0.295299064      0
#> 30    1.336090016      0
#> 31    1.219470522      0
#> 32    0.074061995      0
#> 33    0.709328202      0
#> 34    0.457135984      0
#> 35   -0.988130169      0
#> 36    0.913404301      0
#> 37    1.732321953      0
#> 38    0.099653205      0
#> 39    1.294381543      0
#> 40   -1.097934876      0
#> 41    1.585183873      0
#> 42    0.819938374      0
#> 43   -0.547663704      0
#> 44    0.511675860      0
#> 45   -0.324530954      0
#> 46    0.076887043      0
#> 47    1.070135606      0
#> 48    0.680017730      0
#> 49   -0.440817236      0
#> 50   -0.792667868      0
#> 51   -0.963860859      0
#> 52   -0.594808287      0
#> 53    1.042540798      0
#> 54    0.396185466      0
#> 55   -1.652804818      0
#> 56   -0.154965037      0
#> 57    1.474685524      0
#> 58   -0.225236695      0
#> 59   -0.679375764      0
#> 60   -1.750540907      0
#> 61   -1.198296467      0
#> 62    1.368947920      0
#> 63    0.191871571      0
#> 64   -1.147529245      0
#> 65    0.364158399      0
#> 66    1.661155244      0
#> 67   -2.614146378      0
#> 68    0.799786348      0
#> 69   -1.356666122      0
#> 70   -1.130752820      0
#> 71    1.064500458      0
#> 72    0.027804104      0
#> 73    1.675461685      0
#> 74   -0.729182546      0
#> 75   -0.078608238      0
#> 76    0.200428371      0
#> 77   -0.306301114      0
#> 78    1.459870743      0
#> 79    0.803492704      0
#> 80   -0.385462006      0
#> 81    1.549242365      0
#> 82    0.011690081      0
#> 83   -0.563652605      0
#> 84    1.357484919      0
#> 85    0.525027617      0
#> 86    0.922866947      0
#> 87    0.340238141      0
#> 88   -0.160955885      0
#> 89   -0.934048131      0
#> 90   -0.457879556      0
#> 91    1.618326111      0
#> 92   -0.011627456      0
#> 93    1.355425050      0
#> 94    0.247671759      0
#> 95   -1.337247016      0
#> 96   -0.637530649      0
#> 97   -0.095722499      0
#> 98   -0.404778898      0
#> 99   -0.891196613      0
#> 100  -0.940277287      0
#> 101   0.007885139      0
#> 102  -0.797203201      0
#> 103  -0.854032937      0
#> 104  -1.002149258      0
#> 105  -1.215713894      0
#> 106   0.674684627      0
#> 107  -0.770549532      0
#> 108   0.058873711      0
#> 109   1.342605843      0
#> 110  -2.354314909      0
#> 111  -0.052232939      0
#> 112   0.264076799      0
#> 113  -1.920973383      0
#> 114  -1.553522934      0
#> 115  -1.796528235      0
#> 116  -0.272627489      0
#> 117  -0.625796456      0
#> 118  -0.891575291      0
#> 119   1.268094827      0
#> 120   0.494520880      0
#> 121  -0.265420560      0
#> 122   1.154860955      0
#> 123  -0.022712725      0
#> 124  -0.025755907      0
#> 125   0.279556749      0
#> 126  -0.539955000      0
#> 127  -0.629245390      0
#> 128  -1.673970449      0
#> 129   1.319345427      0
#> 130  -0.875898589      0
#> 131  -0.711440135      0
#> 132   0.850353604      0
#> 133  -0.375262790      0
#> 134   1.000548192      0
#> 135   1.134258765      0
#> 136  -0.342164916      0
#> 137  -0.310520399      0
#> 138   1.201385653      0
#> 139  -0.079922225      0
#> 140  -0.534532696      0
#> 141  -0.134695175      0
#> 142   1.608952531      0
#> 143   1.237544464      0
#> 144   0.620874103      0
#> 145   0.320196678      0
#> 146   0.706368542      0
#> 147  -0.441677983      0
#> 148   0.508676689      0
#> 149  -1.412634973      0
#> 150  -0.032126927      0
#> 151   3.333232085      0
#> 152  -0.908513939      0
#> 153  -1.547431746      0
#> 154   0.697525356      0
#> 155  -1.696096193      0
#> 156  -0.336505766      0
#> 157   0.204603130      0
#> 158   0.311651054      0
#> 159  -1.067470377      0
#> 160   0.883639424      0
#> 161  -0.236122021      0
#> 162   2.052822569      0
#> 163  -0.678527690      0
#> 164  -0.358159926      0
#> 165   0.370709716      0
#> 166  -0.859590072      0
#> 167  -1.520842512      0
#> 168   0.882686880      0
#> 169   0.959746717      0
#> 170  -1.018587895      0
#> 171   1.181863645      0
#> 172  -0.432894667      0
#> 173   0.228065929      0
#> 174   2.043854607      0
#> 175   1.455044739      0
#> 176   1.239094916      0
#> 177   1.745963781      0
#> 178   0.258809894      0
#> 179  -0.016853033      0
#> 180   0.286985846      0
#> 181   0.135289695      0
#> 182   0.031991042      0
#> 183  -1.189001262      0
#> 184  -0.027062110      0
#> 185   1.746383970      0
#> 186  -0.084853421      0
#> 187  -1.039551930      0
#> 188  -0.314426221      0
#> 189   1.021257256      0
#> 190   1.275918118      0
#> 191   2.325603400      0
#> 192  -0.426736927      0
#> 193   0.979363859      0
#> 194   0.010476475      0
#> 195   0.564097556      0
#> 196  -0.159772790      0
#> 197  -0.843436389      0
#> 198  -1.039109508      0
#> 199  -0.558252338      0
#> 200   0.387067559      0
#> 201   0.515885008      0
#> 202   0.085554313      0
#> 203  -1.310693258      0
#> 204   0.910309606      0
#> 205  -0.951048049      0
#> 206   2.146998935      0
#> 207   0.808002586      0
#> 208  -0.056371503      0
#> 209  -0.036397706      0
#> 210   1.027304929      0
#> 211  -0.718713464      0
#> 212  -1.434718654      0
#> 213   0.796244480      0
#> 214   1.030382100      0
#> 215  -0.227346089      0
#> 216   0.884529193      0
#> 217  -0.367874233      0
#> 218  -1.788994689      0
#> 219  -0.861059769      0
#> 220  -0.608293253      0
#> 221   1.161425688      0
#> 222   0.847386239      0
#> 223  -1.409961533      0
#> 224  -0.503136228      0
#> 225   1.461893716      0
#> 226  -0.861162287      0
#> 227   0.026749299      0
#> 228  -0.379179132      0
#> 229  -0.879574398      0
#> 230  -1.402978001      0
#> 231  -1.475652330      0
#> 232   1.451640228      0
#> 233  -0.835933317      0
#> 234   0.901302006      0
#> 235  -0.157682466      0
#> 236  -0.221335603      0
#> 237   1.426236499      0
#> 238  -1.247310225      0
#> 239   0.284410106      0
#> 240   1.296206251      0
#> 241   0.671540544      0
#> 242   1.080120949      0
#> 243  -1.558859139      0
#> 244   0.623617380      0
#> 245   1.226750768      0
#> 246  -0.525966844      0
#> 247   1.816002572      0
#> 248  -0.686910221      0
#> 249   0.454510527      0
#> 250   0.133036393      0
#> 251  -0.321776626      0
#> 252   1.779131415      0
#> 253   0.908876666      0
#> 254   1.027785666      0
#> 255   1.381088323      0
#> 256  -0.785166042      0
#> 257  -0.262397976      0
#> 258  -0.421212273      0
#> 259  -0.734199033      0
#> 260   1.399038069      0
#> 261  -0.836795728      0
#> 262  -0.205444215      0
#> 263  -1.300189956      0
#> 264  -0.033151308      0
#> 265  -1.097747897      0
#> 266   0.216944250      0
#> 267   0.561845168      0
#> 268   0.431373243      0
#> 269  -0.125372993      0
#> 270  -1.341825963      0
#> 271  -0.607389191      0
#> 272   0.639364103      0
#> 273   0.441353941      0
#> 274  -0.146441418      0
#> 275   1.538825547      0
#> 276  -0.380466627      0
#> 277  -1.342834642      0
#> 278  -0.586325074      0
#> 279  -0.733295984      0
#> 280  -0.341173682      0
#> 281   1.359782757      0
#> 282  -0.555259545      0
#> 283   0.179379755      0
#> 284  -0.339100932      0
#> 285  -2.798235887      0
#> 286  -1.450280095      0
#> 287  -0.219419919      0
#> 288   2.338366251      0
#> 289   0.344229678      0
#> 290   0.060311443      0
#> 291  -0.959849830      0
#> 292  -0.874354831      0
#> 293   0.141391038      0
#> 294  -0.675310798      0
#> 295   1.315048056      0
#> 296   0.066897381      0
#> 297   0.820099092      0
#> 298   0.445522113      0
#> 299   1.597875370      0
#> 300   1.753930492      0
#> 301  -0.415274705      0
#> 302   1.165431738      0
#> 303   0.814915312      0
#> 304   0.890886811      0
#> 305   1.110019877      0
#> 306   0.753954618      0
#> 307   0.187128925      0
#> 308  -0.942059246      0
#> 309   0.383681107      0
#> 310   1.257315911      0
#> 311  -1.702126992      0
#> 312  -0.044398658      0
#> 313   0.339936987      0
#> 314  -1.186410297      0
#> 315   1.037607498      0
#> 316   1.692116878      0
#> 317  -0.895672816      0
#> 318   0.116959556      0
#> 319  -0.117609654      0
#> 320  -3.666633440      0
#> 321  -0.306574279      0
#> 322   0.991961616      0
#> 323  -2.313608692      0
#> 324   0.270945894      0
#> 325   1.808842268      0
#> 326  -0.487819795      0
#> 327  -0.672427181      0
#> 328   0.005765249      0
#> 329   0.584421871      0
#> 330   1.226079636      0
#> 331   0.812838279      0
#> 332   0.474657109      0
#> 333  -0.576770790      0
#> 334  -1.621836783      0
#> 335  -0.310531976      0
#> 336   0.510764303      0
#> 337  -1.258131861      0
#> 338   0.366506544      0
#> 339   1.735864203      0
#> 340   1.043101891      0
#> 341  -0.455185747      0
#> 342   1.125768003      0
#> 343  -0.388806079      0
#> 344  -2.139542946      0
#> 345   0.221870759      0
#> 346   1.381935520      0
#> 347  -0.711355854      0
#> 348  -1.347962521      0
#> 349   0.097035134      0
#> 350  -1.582251864      0
#> 351  -1.543429308      0
#> 352  -0.263348423      0
#> 353   1.683263135      0
#> 354   0.536997918      0
#> 355   0.663224567      0
#> 356  -0.717709269      0
#> 357   0.870345765      0
#> 358   0.818531334      0
#> 359   1.052345469      0
#> 360   0.898437796      0
#> 361   1.445777465      0
#> 362   1.737140107      0
#> 363  -1.666057866      0
#> 364   0.188452242      0
#> 365   0.327219965      0
#> 366  -0.210091770      0
#> 367  -0.893709763      0
#> 368  -0.085553715      0
#> 369  -0.273757429      0
#> 370  -2.122695528      0
#> 371   0.768427321      0
#> 372   1.356480471      0
#> 373   0.571095645      0
#> 374   1.668033228      0
#> 375  -0.922930009      0
#> 376  -0.244792801      0
#> 377  -1.908533256      0
#> 378  -1.369292903      0
#> 379  -1.165555796      0
#> 380   0.001221567      0
#> 381   1.015357765      0
#> 382   0.068293087      0
#> 383  -2.388246740      0
#> 384  -0.384451431      0
#> 385  -1.801417065      0
#> 386  -1.831384041      0
#> 387   0.110893131      0
#> 388  -1.712135163      0
#> 389   1.364010092      0
#> 390  -0.087634017      0
#> 391  -0.371731733      0
#> 392  -1.748083726      0
#> 393  -0.744250787      0
#> 394   0.255881477      0
#> 395  -0.687200266      0
#> 396  -1.194814672      0
#> 397  -1.590284213      0
#> 398   0.771572342      0
#> 399   0.752333262      0
#> 400  -0.235379150      0
#> 401   0.038293085      0
#> 402  -0.131169837      0
#> 403   0.867806149      0
#> 404  -0.294730929      0
#> 405  -2.047242266      0
#> 406  -1.027313242      0
#> 407  -0.343741663      0
#> 408  -0.450855615      0
#> 409   1.058099230      0
#> 410  -1.263906912      0
#> 411  -1.661083110      0
#> 412   0.320445541      0
#> 413  -0.411608676      0
#> 414   0.980158261      0
#> 415  -0.261594552      0
#> 416  -0.126637561      0
#> 417  -0.832953730      0
#> 418   0.211159444      0
#> 419   0.457894383      0
#> 420  -0.323790455      0
#> 421   1.037684609      0
#> 422  -0.256035236      0
#> 423   0.534539066      0
#> 424  -0.256113355      0
#> 425   1.578847258      0
#> 426   0.877028018      0
#> 427   1.006960048      0
#> 428   0.863264418      0
#> 429   0.209216621      0
#> 430  -0.781185648      0
#> 431  -0.299213161      0
#> 432  -1.742282969      0
#> 433   1.470590285      0
#> 434  -0.843668998      0
#> 435   1.891441606      0
#> 436  -0.192799925      0
#> 437   0.840636319      0
#> 438  -1.183046734      0
#> 439   2.185865993      0
#> 440  -0.519631323      0
#> 441  -0.451533742      0
#> 442   1.596132632      0
#> 443  -0.437681852      0
#> 444   0.160539064      0
#> 445   0.474234830      0
#> 446  -0.064103006      0
#> 447  -0.699439869      0
#> 448   0.810235067      0
#> 449   1.536212195      0
#> 450   0.647017266      0
#> 451   0.171716272      0
#> 452   1.628874631      0
#> 453  -1.028656370      0
#> 454   1.107265952      0
#> 455  -0.514482675      0
#> 456  -1.795259543      0
#> 457  -0.286855581      0
#> 458  -0.352409310      0
#> 459  -3.540772974      0
#> 460  -0.008995235      0
#> 461  -0.359375321      0
#> 462   1.583896468      0
#> 463   0.049162994      0
#> 464  -0.256260229      0
#> 465  -1.093219064      0
#> 466   1.127636373      0
#> 467  -0.736583465      0
#> 468   0.315551330      0
#> 469  -1.011216966      0
#> 470   0.945870615      0
#> 471   0.610269065      0
#> 472  -1.064043631      0
#> 473   0.078135986      0
#> 474  -1.059418848      0
#> 475  -0.011995788      0
#> 476  -0.082229365      0
#> 477  -0.414935556      0
#> 478   1.986705286      0
#> 479   0.932809441      0
#> 480   1.017026624      0
#> 481  -2.249728616      0
#> 482   0.621193906      0
#> 483  -1.117287497      0
#> 484   0.442075471      0
#> 485  -1.316909198      0
#> 486   1.808712042      0
#> 487  -0.206976984      0
#> 488  -0.978611878      0
#> 489  -0.422332363      0
#> 490  -1.062368878      0
#> 491   0.763098481      0
#> 492   0.833091662      0
#> 493   1.478574328      0
#> 494  -0.202820689      0
#> 495  -0.169157803      0
#> 496  -1.053113827      0
#> 497   1.674738547      0
#> 498   0.097874618      0
#> 499  -0.027121066      0
#> 500  -2.258257990      0
#> 501  -0.249690542      0
#> 502  -0.540824840      0
#> 503  -0.873691336      0
#> 504  -0.164386698      0
#> 505   1.780481428      0
#> 506  -0.964244206      0
#> 507   0.130971071      0
#> 508   1.960874789      0
#> 509  -0.673958833      0
#> 510  -0.615626200      0
#> 511  -0.472867238      0
#> 512  -0.838840666      0
#> 513  -0.158688318      0
#> 514   0.389574346      0
#> 515  -2.000902791      0
#> 516   2.001515545      0
#> 517  -1.155805017      0
#> 518  -0.908546562      0
#> 519  -0.359290652      0
#> 520   0.580669161      0
#> 521  -0.483267864      0
#> 522   0.319334382      0
#> 523  -0.798724795      0
#> 524   1.434701339      0
#> 525   0.221909851      0
#> 526  -1.929878684      0
#> 527   1.058050977      0
#> 528   1.624627586      0
#> 529  -0.790389592      0
#> 530  -1.550935965      0
#> 531   0.308776036      0
#> 532   1.620961300      0
#> 533  -0.567585025      0
#> 534   0.568697617      0
#> 535   0.833237716      0
#> 536   1.754471099      0
#> 537   1.679808430      0
#> 538   0.756495499      0
#> 539   0.594887601      0
#> 540  -1.598848725      0
#> 541  -0.133337729      0
#> 542   1.116261016      0
#> 543   0.401093323      0
#> 544  -0.618444095      0
#> 545   0.205007273      0
#> 546  -0.281239690      0
#> 547  -0.213865525      0
#> 548   0.353752507      0
#> 549   1.284001227      0
#> 550   0.414700272      0
#> 551   0.354236033      0
#> 552   0.515315976      0
#> 553  -0.112414599      0
#> 554  -0.040651831      0
#> 555  -0.810777970      0
#> 556  -1.407046312      0
#> 557  -0.463756053      0
#> 558  -2.518830741      0
#> 559  -2.039271276      0
#> 560   1.356422790      0
#> 561   0.785015868      0
#> 562  -0.295905834      0
#> 563  -0.514026783      0
#> 564  -1.255539874      0
#> 565   0.091187865      0
#> 566   1.159285056      0
#> 567  -0.714564229      0
#> 568   2.537969248      0
#> 569  -1.259671835      0
#> 570   1.181072129      0
#> 571   0.232273914      0
#> 572   0.138189135      0
#> 573  -0.579816919      0
#> 574   1.231455114      0
#> 575  -1.186420885      0
#> 576   0.350360585      0
#> 577   0.761929992      0
#> 578   1.436329952      0
#> 579   0.660507179      0
#> 580  -0.729177476      0
#> 581  -0.019591293      0
#> 582  -0.103277335      0
#> 583  -1.027802036      0
#> 584  -0.187649135      0
#> 585  -0.083550804      0
#> 586  -0.660483995      0
#> 587  -0.901368963      0
#> 588   1.930512971      0
#> 589   0.340128074      0
#> 590   0.119509574      0
#> 591   0.253739016      0
#> 592   1.114041292      0
#> 593   0.095845479      0
#> 594  -0.348195737      0
#> 595   0.312645753      0
#> 596   0.781004397      0
#> 597   1.989603566      0
#> 598  -0.491061509      0
#> 599   2.692190135      0
#> 600   0.106827829      0
#> 601  -2.257199794      0
#> 602  -1.344089693      0
#> 603   0.036750662      0
#> 604   0.340673683      0
#> 605   1.262446065      0
#> 606  -0.944891501      0
#> 607  -0.285146496      0
#> 608  -1.008069096      0
#> 609  -0.832595518      0
#> 610   1.288967893      0
#> 611   0.218658470      0
#> 612  -0.489939208      0
#> 613  -0.785409746      0
#> 614   0.713031674      0
#> 615   0.199362518      0
#> 616  -0.163443817      0
#> 617   1.588947650      0
#> 618   0.607904709      0
#> 619   1.642587519      0
#> 620  -0.597487563      0
#> 621   0.480625927      0
#> 622   1.588177464      0
#> 623   0.425906789      0
#> 624  -0.534869961      0
#> 625  -0.008931342      0
#> 626   1.002378652      0
#> 627   1.493655821      0
#> 628  -0.590373088      0
#> 629   0.617355377      0
#> 630   0.755637221      0
#> 631  -0.370549076      0
#> 632  -1.327864694      0
#> 633  -0.649455066      0
#> 634   1.917115168      0
#> 635   0.717584870      0
#> 636   1.249800985      0
#> 637   0.051777924      0
#> 638  -0.277128744      0
#> 639   0.520022973      0
#> 640  -1.737696816      0
#> 641  -0.106133505      0
#> 642   0.058500181      0
#> 643   0.810121305      0
#> 644  -1.209101133      0
#> 645  -1.500819414      0
#> 646  -1.555776292      0
#> 647  -0.138932200      0
#> 648   0.789016351      0
#> 649   0.201449870      0
#> 650  -2.127329586      0
#> 651  -0.590023696      0
#> 652  -0.402652877      0
#> 653   1.658215114      0
#> 654   1.420777361      0
#> 655   0.676454995      0
#> 656  -0.253863600      0
#> 657   0.322502271      0
#> 658  -0.444206822      0
#> 659  -0.277978040      0
#> 660   0.266895828      0
#> 661   0.898387100      0
#> 662   0.374595992      0
#> 663  -1.352743362      0
#> 664   0.430514131      0
#> 665   1.255336435      0
#> 666   0.340297345      0
#> 667  -0.208224518      0
#> 668  -0.731660207      0
#> 669  -0.025349961      0
#> 670  -0.852501862      0
#> 671  -1.617325077      0
#> 672  -0.182833197      0
#> 673   0.639681703      0
#> 674   1.311428237      0
#> 675  -0.907095263      0
#> 676   0.607281636      0
#> 677   0.857362709      0
#> 678  -0.895487209      0
#> 679  -0.264017595      0
#> 680  -1.898356821      0
#> 681  -1.126362947      0
#> 682  -1.494702031      0
#> 683  -0.413531866      0
#> 684   0.820427315      0
#> 685   0.950398841      0
#> 686  -0.775376814      0
#> 687   1.595620375      0
#> 688  -0.303793712      0
#> 689  -1.016855020      0
#> 690  -1.356500124      0
#> 691   1.508117831      0
#> 692   1.360816652      0
#> 693   0.163949449      0
#> 694   0.026820101      0
#> 695   1.387750557      0
#> 696   0.705439001      0
#> 697  -1.458104750      0
#> 698  -0.052796472      0
#> 699  -0.720139472      0
#> 700   1.079471216      0
#> 701  -2.076566437      0
#> 702  -0.027736144      0
#> 703  -0.135542721      0
#> 704   0.779933011      0
#> 705   0.798087426      0
#> 706  -0.536088932      0
#> 707   1.761694493      0
#> 708  -0.353049072      0
#> 709   1.075255304      0
#> 710   1.744345291      0
#> 711   0.340004559      0
#> 712   1.555215338      0
#> 713   1.689919154      0
#> 714   0.900101175      0
#> 715  -0.217797147      0
#> 716  -0.569741168      0
#> 717   2.195015421      0
#> 718  -0.618803500      0
#> 719   0.384983881      0
#> 720  -1.150604914      0
#> 721   1.226190676      0
#> 722   1.290417398      0
#> 723  -0.407746880      0
#> 724  -0.570383496      0
#> 725  -0.446494667      0
#> 726   0.860935204      0
#> 727   0.436794273      0
#> 728   0.603468882      0
#> 729  -0.404349497      0
#> 730  -0.626692592      0
#> 731   0.319484619      0
#> 732   0.375380572      0
#> 733  -0.216389490      0
#> 734   0.655181901      0
#> 735   1.033918877      0
#> 736   0.691943049      0
#> 737   0.877609805      0
#> 738  -0.190428741      0
#> 739  -1.531792295      0
#> 740  -0.840909324      0
#> 741  -2.168704777      0
#> 742  -0.878234525      0
#> 743   1.834629147      0
#> 744   0.626197302      0
#> 745   0.223714095      0
#> 746  -2.145487453      0
#> 747   0.073573278      0
#> 748   0.810386009      0
#> 749   1.453236698      0
#> 750   0.072212237      0
#> 751  -1.057232571      0
#> 752  -1.281773984      0
#> 753   0.980546882      0
#> 754   0.174799534      0
#> 755  -0.665295192      0
#> 756  -0.967749475      0
#> 757  -2.138207010      0
#> 758   0.638770214      0
#> 759  -1.204618805      0
#> 760  -0.490827383      0
#> 761  -0.529513124      0
#> 762  -1.947757556      0
#> 763  -1.224028777      0
#> 764   0.740523649      0
#> 765  -0.419059800      0
#> 766  -0.673589901      0
#> 767  -0.918679207      0
#> 768  -0.239895649      0
#> 769   1.067487817      0
#> 770  -0.052714929      0
#> 771   1.460132166      0
#> 772  -0.667837657      0
#> 773   1.807961233      0
#> 774  -0.517634678      0
#> 775   1.016827904      0
#> 776   1.925400915      0
#> 777  -1.613045013      0
#> 778   1.175454203      0
#> 779  -0.784093650      0
#> 780   0.890930865      0
#> 781  -0.804839087      0
#> 782  -0.826330279      0
#> 783   0.076354278      0
#> 784   1.834264729      0
#> 785  -0.796874273      0
#> 786  -0.232952463      0
#> 787  -0.967024567      0
#> 788  -0.423461093      0
#> 789  -0.281291782      0
#> 790   0.813005801      0
#> 791  -0.421044795      0
#> 792   1.349000151      0
#> 793  -2.576648725      0
#> 794  -0.157587342      0
#> 795  -1.698497860      0
#> 796  -0.704449468      0
#> 797   0.607802540      0
#> 798   0.115563132      0
#> 799   0.362454896      0
#> 800  -0.081686286      0
#> 801  -0.105981946      0
#> 802  -0.095142257      0
#> 803   2.208474352      0
#> 804   1.671812670      0
#> 805  -1.007258961      0
#> 806  -0.312869385      0
#> 807  -0.298921056      0
#> 808  -0.052069521      0
#> 809   0.994678015      0
#> 810  -0.540474074      0
#> 811  -0.281681367      0
#> 812  -0.503736790      0
#> 813   1.749593861      0
#> 814  -0.681316976      0
#> 815   2.108582272      0
#> 816   0.019452189      0
#> 817  -0.203298703      0
#> 818   1.572597389      0
#> 819   1.353345891      0
#> 820   0.243418149      0
#> 821   0.459505564      0
#> 822  -1.604066318      0
#> 823  -0.247515181      0
#> 824   0.209985935      0
#> 825  -0.152605494      0
#> 826   0.375079119      0
#> 827   1.404569713      0
#> 828  -0.467765824      0
#> 829   0.307547514      0
#> 830  -1.406257699      0
#> 831   0.786473913      0
#> 832  -0.733720820      0
#> 833   0.266021668      0
#> 834   0.280719243      0
#> 835  -0.606641904      0
#> 836   1.747850122      0
#> 837  -0.753080057      0
#> 838  -0.329801238      0
#> 839  -0.345509096      0
#> 840  -1.314401301      0
#> 841  -1.907418682      0
#> 842  -0.452948070      0
#> 843   0.057400381      0
#> 844  -0.312278324      0
#> 845  -1.563345656      0
#> 846  -0.292948403      0
#> 847  -0.782959847      0
#> 848   0.176079866      0
#> 849   1.428306749      0
#> 850   2.484565647      0
#> 851   0.664960000      0
#> 852  -0.822565383      0
#> 853  -0.303105098      0
#> 854   2.622190938      0
#> 855  -0.495379120      0
#> 856   0.556680879      0
#> 857   0.511522952      0
#> 858   0.628428538      0
#> 859  -0.902412220      0
#> 860   1.337180219      0
#> 861  -1.035635257      0
#> 862   0.933109538      0
#> 863  -0.113595790      0
#> 864   0.228205246      0
#> 865  -1.756386584      0
#> 866  -0.133666198      0
#> 867   0.699865053      0
#> 868   0.026885570      0
#> 869   0.104650825      0
#> 870  -0.034769640      0
#> 871  -0.840481327      0
#> 872  -1.189582343      0
#> 873  -0.102527270      0
#> 874   0.630942637      0
#> 875   0.681623038      0
#> 876   0.367287598      0
#> 877   0.277109909      0
#> 878  -0.064391318      0
#> 879  -2.789443694      0
#> 880   0.236808040      0
#> 881   1.083346209      0
#> 882   0.459273089      0
#> 883  -0.296759981      0
#> 884   0.419460252      0
#> 885  -0.208652422      0
#> 886  -0.365002758      0
#> 887  -2.271661491      0
#> 888   0.897456862      0
#> 889  -0.056758332      0
#> 890   1.376480225      0
#> 891  -0.006228466      0
#> 892   0.074092315      0
#> 893  -0.230438012      0
#> 894  -1.152111332      0
#> 895  -0.578903366      0
#> 896  -0.409649296      0
#> 897   0.820319228      0
#> 898   0.086888404      0
#> 899  -0.862558344      0
#> 900  -0.629661174      0
#> 901   2.294962342      1
#> 902   0.831508786      1
#> 903  -2.246877129      1
#> 904  -0.143683644      1
#> 905  -0.365902224      1
#> 906  -1.303344615      1
#> 907  -1.653654862      1
#> 908   1.440876497      1
#> 909   1.720971706      1
#> 910   0.268966531      1
#> 911  -0.523782298      1
#> 912   0.384483739      1
#> 913  -0.220153852      1
#> 914   0.618451821      1
#> 915  -0.471981360      1
#> 916   0.950392083      1
#> 917   0.641467333      1
#> 918  -0.612519018      1
#> 919   0.712019267      1
#> 920   0.440799427      1
#> 921   1.398108377      1
#> 922   0.521476345      1
#> 923   0.180473347      1
#> 924   1.717164057      1
#> 925   0.926907236      1
#> 926   2.669396990      1
#> 927  -0.618787162      1
#> 928  -1.925358610      1
#> 929  -0.266471330      1
#> 930   0.650149622      1
#> 931  -0.286559288      1
#> 932   1.264327765      1
#> 933   1.371349609      1
#> 934   0.344818791      1
#> 935   1.885977473      1
#> 936   0.588345508      1
#> 937   0.414993845      1
#> 938   1.266787389      1
#> 939   3.159026520      1
#> 940   1.410881489      1
#> 941  -0.636258616      1
#> 942  -0.581752949      1
#> 943   0.297761197      1
#> 944   1.793937757      1
#> 945   0.399503555      1
#> 946   0.833378169      1
#> 947   0.872341795      1
#> 948   0.740479196      1
#> 949   1.355000321      1
#> 950   1.264089192      1
#> 951  -0.320117967      1
#> 952   1.230240462      1
#> 953   0.696540842      1
#> 954   0.497770506      1
#> 955   2.568666219      1
#> 956   1.479963508      1
#> 957   1.309750159      1
#> 958   1.221473486      1
#> 959  -0.051921457      1
#> 960   2.039237670      1
#> 961   0.176233092      1
#> 962   0.406073155      1
#> 963   3.033087881      1
#> 964   0.064985275      1
#> 965   2.816749718      1
#> 966  -0.094809009      1
#> 967  -0.602264689      1
#> 968   1.447940666      1
#> 969   0.526164022      1
#> 970  -0.113649881      1
#> 971   3.215683521      1
#> 972   2.210655373      1
#> 973   0.089279883      1
#> 974   1.092310257      1
#> 975  -0.491730712      1
#> 976   1.087287350      1
#> 977  -0.716396988      1
#> 978   2.215316010      1
#> 979  -0.287620840      1
#> 980  -0.242913835      1
#> 981  -0.502859918      1
#> 982   2.200146654      1
#> 983  -0.657623752      1
#> 984   1.670179168      1
#> 985   1.325872416      1
#> 986   1.104321018      1
#> 987   0.225277785      1
#> 988   1.477643967      1
#> 989   1.335927793      1
#> 990  -0.569278196      1
#> 991  -0.844942790      1
#> 992   0.396833630      1
#> 993   0.825655333      1
#> 994   1.018250422      1
#> 995   0.975796548      1
#> 996  -0.146459606      1
#> 997  -0.691207488      1
#> 998   0.679307698      1
#> 999  -0.074340029      1
#> 1000 -0.430445666      1