Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     1.183200456      0
#> 2     0.880754920      0
#> 3    -0.891160469      0
#> 4     0.544524977      0
#> 5     0.240956910      0
#> 6     0.264412694      0
#> 7    -1.321822402      0
#> 8    -0.646660843      0
#> 9    -1.084283308      0
#> 10    1.086119179      0
#> 11    0.396844426      0
#> 12    0.493599218      0
#> 13    0.668576351      0
#> 14   -0.918476187      0
#> 15   -1.126047025      0
#> 16    0.476373245      0
#> 17   -1.204670263      0
#> 18   -0.430336469      0
#> 19    1.491004278      0
#> 20    1.196774255      0
#> 21   -1.499971787      0
#> 22   -1.660306370      0
#> 23    0.950568782      0
#> 24   -0.438080386      0
#> 25    0.087223145      0
#> 26    0.577939654      0
#> 27    1.126456682      0
#> 28    0.760858014      0
#> 29    0.379942726      0
#> 30   -0.950832524      0
#> 31    0.098924429      0
#> 32    0.951842840      0
#> 33    0.543205786      0
#> 34   -0.044243116      0
#> 35   -0.073809516      0
#> 36   -1.626061550      0
#> 37    1.271499996      0
#> 38    1.680504393      0
#> 39   -0.784596958      0
#> 40   -1.213728343      0
#> 41   -1.655437842      0
#> 42   -1.731719249      0
#> 43    2.866887625      0
#> 44   -0.648909021      0
#> 45    1.601813177      0
#> 46    0.374066956      0
#> 47    0.207798103      0
#> 48    0.328898169      0
#> 49   -0.385284821      0
#> 50   -1.095942019      0
#> 51    0.357787097      0
#> 52    0.224763053      0
#> 53    1.154741928      0
#> 54   -1.308239492      0
#> 55    2.374620795      0
#> 56    0.591090467      0
#> 57    0.636546429      0
#> 58   -0.922508572      0
#> 59    0.894323322      0
#> 60   -0.970471419      0
#> 61    0.183235833      0
#> 62   -0.004390687      0
#> 63    1.041207301      0
#> 64   -1.446230612      0
#> 65    1.404157904      0
#> 66   -0.223850329      0
#> 67    0.280498721      0
#> 68    0.400131383      0
#> 69    2.113378654      0
#> 70    0.932490552      0
#> 71   -0.176697189      0
#> 72    0.352780999      0
#> 73   -0.753710390      0
#> 74   -0.014221762      0
#> 75    0.506019645      0
#> 76    0.759003191      0
#> 77    1.175197101      0
#> 78   -0.351417613      0
#> 79   -0.285303345      0
#> 80    1.194140870      0
#> 81   -0.550071158      0
#> 82   -0.161615411      0
#> 83   -2.474664192      0
#> 84    0.054397014      0
#> 85    0.475416207      0
#> 86    1.391113880      0
#> 87   -1.878410715      0
#> 88    0.688727462      0
#> 89    1.401808240      0
#> 90    0.485159485      0
#> 91   -1.628820205      0
#> 92   -0.433526676      0
#> 93   -0.490912665      0
#> 94    0.431612651      0
#> 95    0.176066001      0
#> 96   -1.642798231      0
#> 97    0.360808520      0
#> 98   -2.167784320      0
#> 99   -0.437974974      0
#> 100  -2.286612293      0
#> 101   1.468255858      0
#> 102   0.556897991      0
#> 103   1.866181471      0
#> 104   0.242619003      0
#> 105   0.651164038      0
#> 106   1.015971687      0
#> 107  -0.640651837      0
#> 108  -0.451145187      0
#> 109  -2.357080840      0
#> 110   0.892170291      0
#> 111  -1.987749569      0
#> 112   0.249419861      0
#> 113  -0.292417438      0
#> 114   0.455062617      0
#> 115  -1.161165599      0
#> 116  -1.340579643      0
#> 117  -0.959212523      0
#> 118   1.014364084      0
#> 119  -1.866984869      0
#> 120   0.144689280      0
#> 121  -1.701883353      0
#> 122   0.625579960      0
#> 123  -0.533399644      0
#> 124  -1.884646980      0
#> 125   0.443134929      0
#> 126  -2.143683262      0
#> 127  -1.101917942      0
#> 128   0.037786003      0
#> 129  -1.815848593      0
#> 130  -0.973537740      0
#> 131   1.259428219      0
#> 132   0.951398867      0
#> 133   1.004388743      0
#> 134  -0.213908870      0
#> 135   0.550602604      0
#> 136   0.095601849      0
#> 137   0.687897256      0
#> 138   0.868781199      0
#> 139  -0.806184621      0
#> 140   2.232792388      0
#> 141  -0.787489152      0
#> 142  -0.510199421      0
#> 143   0.097150928      0
#> 144  -1.763326609      0
#> 145   1.179563178      0
#> 146   1.437190834      0
#> 147   0.041187367      0
#> 148  -0.459178668      0
#> 149   1.300006588      0
#> 150  -0.952582955      0
#> 151   0.127443267      0
#> 152   0.113006484      0
#> 153  -0.264182656      0
#> 154   2.772719893      0
#> 155  -1.594892969      0
#> 156   1.283716709      0
#> 157   1.150300698      0
#> 158  -0.301775510      0
#> 159  -0.975421769      0
#> 160  -0.467002210      0
#> 161  -0.732447892      0
#> 162  -1.761523256      0
#> 163  -1.904996490      0
#> 164   0.299393851      0
#> 165   0.310418452      0
#> 166  -0.874041847      0
#> 167  -0.961234531      0
#> 168   0.379458053      0
#> 169  -0.713890228      0
#> 170  -1.131641793      0
#> 171  -0.152852296      0
#> 172   0.408647109      0
#> 173   1.423612188      0
#> 174   0.342137184      0
#> 175   0.203511733      0
#> 176   1.234131853      0
#> 177  -0.751528447      0
#> 178  -0.090467919      0
#> 179  -0.616095880      0
#> 180   0.735791906      0
#> 181  -2.272971006      0
#> 182   1.157506184      0
#> 183   0.251990238      0
#> 184  -0.666838989      0
#> 185   0.488181459      0
#> 186   1.676936901      0
#> 187  -0.888013319      0
#> 188   0.498458730      0
#> 189   0.580252245      0
#> 190   0.991866325      0
#> 191   0.906267318      0
#> 192   0.104680397      0
#> 193   0.510625852      0
#> 194   0.238546788      0
#> 195   1.306661861      0
#> 196   1.083696815      0
#> 197   0.539573961      0
#> 198   0.943748293      0
#> 199   1.529029529      0
#> 200   0.315628481      0
#> 201  -0.338399577      0
#> 202  -0.254421937      0
#> 203  -0.754837160      0
#> 204   1.147364641      0
#> 205   0.054347911      0
#> 206   0.032386125      0
#> 207  -1.322049622      0
#> 208   0.776344560      0
#> 209  -0.340964809      0
#> 210  -2.006652668      0
#> 211  -0.637059797      0
#> 212  -1.342523425      0
#> 213  -0.381440871      0
#> 214   0.226875391      0
#> 215   1.757309271      0
#> 216   1.123611678      0
#> 217  -0.504305768      0
#> 218  -1.231100089      0
#> 219  -0.724117371      0
#> 220   0.909368866      0
#> 221   1.404040590      0
#> 222  -1.816942153      0
#> 223  -0.639329063      0
#> 224   1.204280789      0
#> 225  -1.285373296      0
#> 226   0.841251076      0
#> 227  -0.693722841      0
#> 228   1.494231258      0
#> 229   1.464650039      0
#> 230  -0.144016296      0
#> 231  -0.181675108      0
#> 232   0.514800398      0
#> 233  -0.077250942      0
#> 234   0.151748098      0
#> 235   2.211562733      0
#> 236  -0.833949544      0
#> 237  -1.308214843      0
#> 238   0.625216037      0
#> 239  -2.073754751      0
#> 240  -0.296833200      0
#> 241  -1.942963549      0
#> 242  -0.742774937      0
#> 243   1.271220605      0
#> 244  -2.062597797      0
#> 245   0.677703942      0
#> 246  -0.022324062      0
#> 247   0.696277343      0
#> 248   1.888428070      0
#> 249   0.110840102      0
#> 250   1.175904590      0
#> 251   1.590002170      0
#> 252   0.387763132      0
#> 253   0.243411720      0
#> 254  -0.238724986      0
#> 255   0.622916585      0
#> 256  -0.884370637      0
#> 257  -1.820526062      0
#> 258   1.578921654      0
#> 259   0.350362263      0
#> 260  -0.176954120      0
#> 261   1.600912742      0
#> 262   0.073621777      0
#> 263  -0.513572339      0
#> 264  -0.681238687      0
#> 265   1.102335519      0
#> 266  -1.085426087      0
#> 267   0.122237370      0
#> 268  -0.713332540      0
#> 269  -1.593371450      0
#> 270   0.268205246      0
#> 271  -1.983346449      0
#> 272   0.713719104      0
#> 273  -0.541906175      0
#> 274   2.131407009      0
#> 275   1.089034325      0
#> 276   2.173705348      0
#> 277   1.352938699      0
#> 278  -1.280081592      0
#> 279  -0.358519089      0
#> 280  -2.838708572      0
#> 281  -1.014528210      0
#> 282   0.566978764      0
#> 283  -2.300423074      0
#> 284   0.557642099      0
#> 285  -1.937797547      0
#> 286   1.054986720      0
#> 287  -0.990378575      0
#> 288   0.186473184      0
#> 289   0.432765427      0
#> 290  -2.095040061      0
#> 291  -0.022700206      0
#> 292  -1.030311569      0
#> 293  -0.559266074      0
#> 294  -0.512953038      0
#> 295  -1.822055052      0
#> 296  -0.887045118      0
#> 297  -2.513280895      0
#> 298  -1.443506684      0
#> 299  -0.114540470      0
#> 300  -2.030441156      0
#> 301  -0.090064376      0
#> 302   0.613393133      0
#> 303   0.731768878      0
#> 304   0.796502829      0
#> 305  -1.101447957      0
#> 306   0.440272713      0
#> 307   0.852393602      0
#> 308  -1.218796437      0
#> 309   0.976415537      0
#> 310   0.435164390      0
#> 311  -0.396984255      0
#> 312   0.099310411      0
#> 313   0.438311354      0
#> 314   2.411493600      0
#> 315   0.668228172      0
#> 316  -0.400964451      0
#> 317   0.714573567      0
#> 318  -0.868221108      0
#> 319  -0.499934272      0
#> 320   1.796566057      0
#> 321  -1.426660604      0
#> 322   0.146861756      0
#> 323   0.149704466      0
#> 324  -0.014280625      0
#> 325   0.634470936      0
#> 326  -1.473314455      0
#> 327   0.450186719      0
#> 328  -0.908930444      0
#> 329   0.085446970      0
#> 330  -0.373890261      0
#> 331   0.487395869      0
#> 332  -1.479183901      0
#> 333   0.581818606      0
#> 334  -0.164403235      0
#> 335   0.090935413      0
#> 336  -0.644203435      0
#> 337   0.418034383      0
#> 338  -0.459530683      0
#> 339  -2.135203074      0
#> 340   0.166074060      0
#> 341   0.072814015      0
#> 342   0.277382740      0
#> 343  -0.371788962      0
#> 344  -0.191482674      0
#> 345  -0.829861760      0
#> 346   0.332262376      0
#> 347  -0.085208459      0
#> 348  -1.782220538      0
#> 349  -0.957698051      0
#> 350  -1.562714169      0
#> 351  -0.189683340      0
#> 352   0.112592300      0
#> 353   0.796533072      0
#> 354   0.322058399      0
#> 355  -0.148495873      0
#> 356  -0.769141581      0
#> 357   0.926012865      0
#> 358  -0.012843118      0
#> 359  -1.362051798      0
#> 360  -0.566309801      0
#> 361  -0.178064356      0
#> 362  -1.692042985      0
#> 363   0.795996545      0
#> 364   0.061240012      0
#> 365   2.311755723      0
#> 366   0.414723162      0
#> 367  -2.348702592      0
#> 368  -0.034347313      0
#> 369  -1.662937763      0
#> 370  -0.963888903      0
#> 371  -0.924880434      0
#> 372   1.340987817      0
#> 373   0.512697407      0
#> 374  -0.431475297      0
#> 375  -0.518865189      0
#> 376  -0.616337332      0
#> 377  -0.244912165      0
#> 378  -0.332720447      0
#> 379  -0.565056868      0
#> 380   0.825404532      0
#> 381   0.334708798      0
#> 382  -0.012313210      0
#> 383  -1.270663250      0
#> 384   2.232853442      0
#> 385   1.432071305      0
#> 386   0.273998685      0
#> 387  -0.881978697      0
#> 388  -0.401893933      0
#> 389  -0.261259043      0
#> 390   0.318298345      0
#> 391   0.601061128      0
#> 392   1.013528138      0
#> 393   1.200158229      0
#> 394   0.416952283      0
#> 395  -0.610927300      0
#> 396   1.525763751      0
#> 397   0.329216204      0
#> 398  -1.870030609      0
#> 399   0.052169730      0
#> 400  -0.386351876      0
#> 401  -1.476464281      0
#> 402   0.910316785      0
#> 403   0.465811305      0
#> 404  -0.897592939      0
#> 405  -0.512014170      0
#> 406  -2.055371338      0
#> 407   0.648230373      0
#> 408  -0.426190350      0
#> 409   0.947036204      0
#> 410   0.375308810      0
#> 411  -1.773127282      0
#> 412  -0.169652066      0
#> 413  -1.148624189      0
#> 414   1.376394583      0
#> 415   1.595619223      0
#> 416   1.173901901      0
#> 417  -1.281567071      0
#> 418   0.713766552      0
#> 419  -2.256889287      0
#> 420  -1.549966496      0
#> 421   1.126099011      0
#> 422   0.053680497      0
#> 423  -0.180606447      0
#> 424  -1.957425954      0
#> 425  -0.875088419      0
#> 426  -0.834840411      0
#> 427  -0.442654122      0
#> 428  -0.572430560      0
#> 429   0.059739887      0
#> 430   0.834004485      0
#> 431  -1.005604112      0
#> 432  -0.007451250      0
#> 433  -0.842814277      0
#> 434  -0.246399589      0
#> 435  -0.271209864      0
#> 436   0.050396406      0
#> 437  -0.880012067      0
#> 438   0.440654010      0
#> 439  -0.152105121      0
#> 440  -0.045467756      0
#> 441   0.421445921      0
#> 442   0.389326281      0
#> 443  -1.329780907      0
#> 444   1.327890473      0
#> 445   1.247053213      0
#> 446   0.128601577      0
#> 447   1.140891783      0
#> 448   0.943876427      0
#> 449  -1.236075059      0
#> 450   2.340247359      0
#> 451  -0.127932536      0
#> 452   0.970506180      0
#> 453  -1.002558458      0
#> 454   1.471488269      0
#> 455  -0.351496646      0
#> 456   1.092785247      0
#> 457  -1.041689368      0
#> 458   0.181905319      0
#> 459  -0.790778849      0
#> 460   0.313850364      0
#> 461  -0.580674870      0
#> 462   0.672607087      0
#> 463   0.448943462      0
#> 464  -0.426969243      0
#> 465   0.057306554      0
#> 466  -0.019966730      0
#> 467   0.939115974      0
#> 468   1.526644121      0
#> 469   0.539982967      0
#> 470   0.332581181      0
#> 471  -0.101944887      0
#> 472   0.311007379      0
#> 473  -1.005254160      0
#> 474   1.229350714      0
#> 475  -1.613689908      0
#> 476   0.992763189      0
#> 477   0.045872714      0
#> 478   2.301213334      0
#> 479   0.822633934      0
#> 480   0.547750262      0
#> 481   0.332445936      0
#> 482  -0.614635821      0
#> 483   0.527235810      0
#> 484   1.307233470      0
#> 485   1.255280210      0
#> 486   0.275031740      0
#> 487  -0.334689511      0
#> 488   0.463232262      0
#> 489   3.736297484      0
#> 490  -1.043412571      0
#> 491   0.020260019      0
#> 492   1.284253486      0
#> 493  -1.407933266      0
#> 494   0.395436231      0
#> 495   1.112121805      0
#> 496   0.266739317      0
#> 497  -0.464227885      0
#> 498   1.116055661      0
#> 499  -1.070826236      0
#> 500   1.891213002      0
#> 501   1.208526553      0
#> 502  -0.302865526      0
#> 503   0.392664254      0
#> 504  -1.112914965      0
#> 505   0.350549830      0
#> 506   0.664259780      0
#> 507   0.708240912      0
#> 508   2.257378040      0
#> 509  -0.960153975      0
#> 510  -0.787102569      0
#> 511   0.139070590      0
#> 512  -0.588590718      0
#> 513   0.432144990      0
#> 514  -0.412538771      0
#> 515  -0.500252482      0
#> 516  -1.050006125      0
#> 517   0.943117338      0
#> 518   1.423686923      0
#> 519  -0.474344281      0
#> 520  -0.354223767      0
#> 521  -0.357931687      0
#> 522  -0.019208844      0
#> 523   1.191376249      0
#> 524  -1.084628993      0
#> 525  -0.737117519      0
#> 526  -0.268498951      0
#> 527   0.762738420      0
#> 528   0.200225719      0
#> 529  -1.484426101      0
#> 530   0.841544568      0
#> 531  -1.229476569      0
#> 532  -1.876238449      0
#> 533  -0.785750010      0
#> 534   0.427762190      0
#> 535   0.715208979      0
#> 536  -1.538517795      0
#> 537   0.342714143      0
#> 538  -0.907941906      0
#> 539  -0.319651232      0
#> 540   2.263112153      0
#> 541  -1.140918623      0
#> 542  -1.206161101      0
#> 543  -1.110914882      0
#> 544  -2.022560864      0
#> 545   0.538151686      0
#> 546   0.660748333      0
#> 547   0.218170730      0
#> 548  -1.913729394      0
#> 549  -0.758655748      0
#> 550   0.562711072      0
#> 551  -1.212783069      0
#> 552   0.514963171      0
#> 553   2.946774134      0
#> 554   1.115880468      0
#> 555   0.712887047      0
#> 556  -1.125858504      0
#> 557   1.812580160      0
#> 558  -0.585837821      0
#> 559   0.696202205      0
#> 560  -1.822747906      0
#> 561  -1.137222151      0
#> 562  -0.669255631      0
#> 563   1.368509198      0
#> 564   0.783174874      0
#> 565  -1.247926180      0
#> 566   0.400804313      0
#> 567   0.728686541      0
#> 568   0.379283184      0
#> 569  -1.486423515      0
#> 570  -0.118323703      0
#> 571   0.315334851      0
#> 572   2.243195222      0
#> 573   0.170086562      0
#> 574   0.250642372      0
#> 575  -1.840744381      0
#> 576   0.595809329      0
#> 577   0.533381348      0
#> 578  -0.571575682      0
#> 579  -0.998787657      0
#> 580   0.860176490      0
#> 581  -0.160693672      0
#> 582  -0.203561379      0
#> 583   0.545953009      0
#> 584  -1.615602082      0
#> 585   0.252928697      0
#> 586   0.401314894      0
#> 587   0.751230149      0
#> 588   0.681330954      0
#> 589   1.310076709      0
#> 590  -0.871275957      0
#> 591  -0.576980931      0
#> 592  -1.839451462      0
#> 593  -0.537357660      0
#> 594   1.057507877      0
#> 595   0.528129807      0
#> 596  -1.263613134      0
#> 597  -0.027021644      0
#> 598   0.513882164      0
#> 599  -0.619753838      0
#> 600  -0.684944299      0
#> 601  -0.914431271      0
#> 602   1.998703891      0
#> 603   0.550240469      0
#> 604  -1.921800380      0
#> 605   2.342335022      0
#> 606   1.583210358      0
#> 607   1.148717470      0
#> 608  -1.049970138      0
#> 609   0.115655208      0
#> 610   0.220883690      0
#> 611   0.515061083      0
#> 612  -2.714067148      0
#> 613  -0.773749256      0
#> 614  -1.605274617      0
#> 615  -0.501234656      0
#> 616   1.305350341      0
#> 617   0.144624585      0
#> 618  -0.860765241      0
#> 619   0.093552329      0
#> 620   0.835397388      0
#> 621   1.456539538      0
#> 622  -0.751165385      0
#> 623   0.581215288      0
#> 624   0.159755256      0
#> 625  -0.597276012      0
#> 626   1.469238362      0
#> 627  -1.074375639      0
#> 628   0.414191532      0
#> 629  -1.659917645      0
#> 630  -1.808549254      0
#> 631  -1.031484947      0
#> 632  -0.386883272      0
#> 633  -0.495087112      0
#> 634   0.727048244      0
#> 635  -1.415070683      0
#> 636  -0.350525402      0
#> 637   0.300559014      0
#> 638  -0.204954730      0
#> 639   1.610575519      0
#> 640  -1.232065659      0
#> 641  -1.460534291      0
#> 642   1.639448473      0
#> 643  -0.821197905      0
#> 644   0.150815436      0
#> 645  -0.784667514      0
#> 646  -0.331435069      0
#> 647   0.442573809      0
#> 648  -0.731587618      0
#> 649   0.248218526      0
#> 650  -1.196206739      0
#> 651  -1.329891673      0
#> 652   0.809446447      0
#> 653  -0.714443979      0
#> 654   0.496458676      0
#> 655   0.720097306      0
#> 656  -2.173043582      0
#> 657  -0.806065094      0
#> 658   0.007775218      0
#> 659  -0.051732151      0
#> 660   0.758984519      0
#> 661   1.803897991      0
#> 662   0.344442741      0
#> 663  -0.559497956      0
#> 664   0.802165682      0
#> 665   0.239628853      0
#> 666   0.213109415      0
#> 667  -1.029664663      0
#> 668   0.895034874      0
#> 669  -0.036725323      0
#> 670   0.484700547      0
#> 671   2.003395342      0
#> 672   0.112854918      0
#> 673  -1.519613608      0
#> 674   1.552411127      0
#> 675   1.562616181      0
#> 676   1.129726063      0
#> 677  -0.059978544      0
#> 678  -0.495898991      0
#> 679  -0.670247406      0
#> 680  -0.189093806      0
#> 681   2.217388149      0
#> 682  -0.562417957      0
#> 683  -0.108501272      0
#> 684  -0.169313170      0
#> 685  -1.856056133      0
#> 686  -0.274936972      0
#> 687  -0.969352866      0
#> 688   1.066444296      0
#> 689   1.565241397      0
#> 690  -0.584259462      0
#> 691  -0.114128885      0
#> 692  -0.816225453      0
#> 693  -0.409917297      0
#> 694   1.381593674      0
#> 695  -0.651546206      0
#> 696  -0.232762333      0
#> 697  -0.743857223      0
#> 698   0.718765164      0
#> 699  -0.573216428      0
#> 700   1.142904325      0
#> 701   1.541322881      0
#> 702   0.879093869      0
#> 703   0.532816631      0
#> 704  -1.097383164      0
#> 705   1.175069559      0
#> 706   0.548612950      0
#> 707  -0.153421454      0
#> 708  -1.180508306      0
#> 709   0.915401248      0
#> 710  -1.056194128      0
#> 711  -1.019297511      0
#> 712  -0.407238778      0
#> 713  -1.201570076      0
#> 714  -0.397221999      0
#> 715   0.795330971      0
#> 716  -0.936203308      0
#> 717   2.033417033      0
#> 718   0.148927139      0
#> 719   1.589785860      0
#> 720   0.674346546      0
#> 721  -0.168222801      0
#> 722  -1.217229420      0
#> 723   0.825450790      0
#> 724   0.228722392      0
#> 725   0.980229077      0
#> 726   0.502059090      0
#> 727   1.682959754      0
#> 728  -0.708714133      0
#> 729  -0.292736737      0
#> 730  -2.179687761      0
#> 731   0.563820648      0
#> 732   0.220344208      0
#> 733  -0.493759558      0
#> 734   1.654239593      0
#> 735   1.058461035      0
#> 736   0.668016797      0
#> 737   1.301012950      0
#> 738  -0.040214759      0
#> 739  -0.819457245      0
#> 740   0.643321093      0
#> 741  -1.112072327      0
#> 742   0.410987132      0
#> 743  -0.118036178      0
#> 744  -0.243106116      0
#> 745   0.754865531      0
#> 746   0.422526337      0
#> 747   0.739002380      0
#> 748  -1.466948326      0
#> 749  -1.771679642      0
#> 750   0.660025980      0
#> 751   0.856734293      0
#> 752  -1.630963720      0
#> 753   1.453720974      0
#> 754   0.250635022      0
#> 755   0.743391817      0
#> 756  -2.688063691      0
#> 757  -0.069844840      0
#> 758  -1.602840553      0
#> 759  -0.459702036      0
#> 760  -0.815960652      0
#> 761   1.330225189      0
#> 762   0.015442173      0
#> 763   0.401860828      0
#> 764   1.529800986      0
#> 765  -0.698259832      0
#> 766   0.976922505      0
#> 767   0.067697831      0
#> 768   0.300108512      0
#> 769   0.221424569      0
#> 770  -0.413924930      0
#> 771   0.121311632      0
#> 772  -0.523072381      0
#> 773  -0.298916739      0
#> 774   0.661136631      0
#> 775   0.743164987      0
#> 776  -0.577033209      0
#> 777  -0.535352849      0
#> 778   1.793626288      0
#> 779  -0.176135518      0
#> 780  -0.459908724      0
#> 781  -0.370809993      0
#> 782   1.596849793      0
#> 783  -0.435103247      0
#> 784   0.255253558      0
#> 785  -0.109833091      0
#> 786  -1.026233794      0
#> 787   0.615079726      0
#> 788   2.554974673      0
#> 789   1.527909954      0
#> 790  -2.123360164      0
#> 791  -0.837357154      0
#> 792   1.015660174      0
#> 793   0.236454262      0
#> 794   0.817688880      0
#> 795   1.235969577      0
#> 796  -0.563173685      0
#> 797  -0.298975905      0
#> 798  -1.176315458      0
#> 799  -1.784936925      0
#> 800  -0.340204432      0
#> 801   0.198181293      0
#> 802   0.979024652      0
#> 803   0.447866416      0
#> 804  -1.582808500      0
#> 805  -0.980310212      0
#> 806   0.972629657      0
#> 807  -2.217835831      0
#> 808  -2.048340490      0
#> 809   0.181691953      0
#> 810  -2.231217279      0
#> 811  -2.097803101      0
#> 812  -0.491769528      0
#> 813  -1.397006163      0
#> 814   0.780396945      0
#> 815  -0.611071454      0
#> 816   1.311433280      0
#> 817  -0.244115896      0
#> 818   0.153152808      0
#> 819   1.432156391      0
#> 820  -0.132507855      0
#> 821  -1.048515109      0
#> 822  -0.054834791      0
#> 823  -0.517510908      0
#> 824  -1.910986919      0
#> 825  -0.132775756      0
#> 826  -0.231982560      0
#> 827   0.741666365      0
#> 828  -0.658346888      0
#> 829  -0.053813371      0
#> 830  -0.983639243      0
#> 831  -1.964956028      0
#> 832  -1.327626533      0
#> 833  -0.316299092      0
#> 834  -0.963753140      0
#> 835  -0.981418351      0
#> 836   0.516331248      0
#> 837   0.546195345      0
#> 838  -1.390705622      0
#> 839  -0.431689674      0
#> 840   0.076172422      0
#> 841  -0.009184771      0
#> 842  -1.157423596      0
#> 843   0.120771088      0
#> 844  -1.030801539      0
#> 845  -0.143906023      0
#> 846  -0.999653985      0
#> 847  -0.772785780      0
#> 848  -0.692428475      0
#> 849  -0.613156658      0
#> 850  -0.866647153      0
#> 851  -1.616622650      0
#> 852   0.303041646      0
#> 853  -1.695613286      0
#> 854  -0.834743113      0
#> 855  -0.419775762      0
#> 856   1.380456234      0
#> 857  -0.143043792      0
#> 858   0.049019870      0
#> 859   0.817198811      0
#> 860   0.606865730      0
#> 861   0.784652003      0
#> 862   0.431670745      0
#> 863  -0.974445653      0
#> 864  -0.344769539      0
#> 865   1.010694733      0
#> 866   1.328066776      0
#> 867  -0.894377356      0
#> 868   0.757929595      0
#> 869   0.949891514      0
#> 870  -0.289807188      0
#> 871   0.426343668      0
#> 872   0.816573449      0
#> 873   0.155992290      0
#> 874   0.612278145      0
#> 875   0.070723940      0
#> 876   1.484171820      0
#> 877  -2.488320853      0
#> 878  -1.298675434      0
#> 879  -0.089909233      0
#> 880   1.370479541      0
#> 881  -0.053095798      0
#> 882  -1.223752635      0
#> 883   1.045714460      0
#> 884  -0.158452645      0
#> 885  -0.782646361      0
#> 886   1.384963878      0
#> 887   0.340021789      0
#> 888  -0.347407481      0
#> 889  -0.100988713      1
#> 890   1.286022904      1
#> 891   1.760353870      1
#> 892   1.600595783      1
#> 893   0.059104959      1
#> 894   0.297556088      1
#> 895   0.973644070      1
#> 896   1.115272536      1
#> 897   0.478688786      1
#> 898  -0.248814792      1
#> 899  -0.673350951      1
#> 900   1.862327333      1
#> 901   0.299431009      1
#> 902   1.512732592      1
#> 903   0.499516505      1
#> 904   1.426806259      1
#> 905   1.288205202      1
#> 906   2.041465628      1
#> 907   0.210304217      1
#> 908   0.754910668      1
#> 909  -0.499412998      1
#> 910   0.285173508      1
#> 911  -0.361719130      1
#> 912   1.649732290      1
#> 913   1.807141383      1
#> 914   1.695675268      1
#> 915   0.150029115      1
#> 916   2.015238681      1
#> 917   0.820339596      1
#> 918  -0.558393473      1
#> 919  -0.268406604      1
#> 920   1.178837627      1
#> 921   2.527310478      1
#> 922   1.701068131      1
#> 923   1.284718569      1
#> 924   0.067869751      1
#> 925   1.751117365      1
#> 926   1.297049699      1
#> 927   0.209998972      1
#> 928   0.921311013      1
#> 929   0.762640791      1
#> 930   1.469976807      1
#> 931   1.769012419      1
#> 932   0.972340495      1
#> 933   0.612731115      1
#> 934   1.273730145      1
#> 935   0.278894519      1
#> 936   1.278381703      1
#> 937   0.421937279      1
#> 938  -0.039270476      1
#> 939   1.387385296      1
#> 940   0.356348199      1
#> 941   1.935040944      1
#> 942   0.716968702      1
#> 943   0.608331378      1
#> 944   0.518830921      1
#> 945   0.015231769      1
#> 946   0.441699364      1
#> 947   2.396573285      1
#> 948   1.395031795      1
#> 949   2.486717252      1
#> 950  -0.229186742      1
#> 951   0.621468673      1
#> 952  -0.245676379      1
#> 953   1.111533104      1
#> 954   1.375281663      1
#> 955   1.984901789      1
#> 956   1.123118223      1
#> 957   0.947403812      1
#> 958   1.043810037      1
#> 959   0.421563382      1
#> 960   1.041157563      1
#> 961   1.197295190      1
#> 962   1.030639436      1
#> 963   1.613429282      1
#> 964  -0.600637085      1
#> 965   2.236983333      1
#> 966   2.321217456      1
#> 967  -0.244165376      1
#> 968   0.370768640      1
#> 969   1.267399189      1
#> 970   1.393531358      1
#> 971   1.297461434      1
#> 972  -0.059733582      1
#> 973   0.266974148      1
#> 974   0.603201334      1
#> 975   1.615905111      1
#> 976   1.665069143      1
#> 977  -1.515683604      1
#> 978   1.634411048      1
#> 979   1.483475360      1
#> 980   2.377629179      1
#> 981   1.564824676      1
#> 982   0.213869035      1
#> 983   1.032427550      1
#> 984   0.784822463      1
#> 985   3.130621016      1
#> 986   0.601251702      1
#> 987   1.500054197      1
#> 988   0.553337023      1
#> 989   1.401563581      1
#> 990  -0.226694692      1
#> 991  -0.894353162      1
#> 992   1.428075601      1
#> 993  -0.373507222      1
#> 994   0.697458629      1
#> 995  -0.203416294      1
#> 996  -0.167232613      1
#> 997  -0.318829901      1
#> 998  -0.768537710      1
#> 999   0.475938056      1
#> 1000  0.925625567      1