Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     1.3297575455      0
#> 2    -0.7364252528      0
#> 3     1.9427457079      0
#> 4     1.2099190582      0
#> 5    -1.1130291216      0
#> 6    -0.2121739585      0
#> 7     0.4887751992      0
#> 8    -0.8188797720      0
#> 9    -0.9573750833      0
#> 10   -0.4370794610      0
#> 11   -0.4227647742      0
#> 12   -0.4594055154      0
#> 13   -0.2071664362      0
#> 14    0.0206080557      0
#> 15   -1.2596750661      0
#> 16    0.8232775417      0
#> 17    0.3500726657      0
#> 18   -0.9917632817      0
#> 19   -0.4255974055      0
#> 20   -1.1674006440      0
#> 21    0.1770239566      0
#> 22    2.5004009995      0
#> 23    0.4229895535      0
#> 24    0.3537557871      0
#> 25    1.5163046191      0
#> 26   -0.6883435217      0
#> 27    1.8797794773      0
#> 28    0.6969762308      0
#> 29   -0.5711413333      0
#> 30   -0.9359662254      0
#> 31   -0.4852947167      0
#> 32    0.4251238769      0
#> 33    2.3137545111      0
#> 34   -0.8648884844      0
#> 35   -2.7156377578      0
#> 36    0.5237630716      0
#> 37    1.4131722598      0
#> 38   -0.4639829615      0
#> 39   -0.0668697070      0
#> 40   -0.4376838155      0
#> 41    1.1275526777      0
#> 42    0.9979302620      0
#> 43   -0.1057869014      0
#> 44    0.1191902221      0
#> 45   -0.3761892230      0
#> 46   -0.6725372017      0
#> 47    1.3439838684      0
#> 48    1.5532688365      0
#> 49   -1.2519205848      0
#> 50    0.5792290913      0
#> 51    1.1678977418      0
#> 52   -0.3324544734      0
#> 53    0.1416126610      0
#> 54   -1.0445281111      0
#> 55   -0.5378371923      0
#> 56   -1.8587672910      0
#> 57   -0.9102733317      0
#> 58    0.6328660491      0
#> 59    0.0137967562      0
#> 60   -1.8643328037      0
#> 61    0.7126494029      0
#> 62   -0.1143386070      0
#> 63    0.3456297494      0
#> 64    0.5096038184      0
#> 65    0.2674551084      0
#> 66   -0.2914953082      0
#> 67    0.3278089393      0
#> 68   -0.0559273738      0
#> 69    0.1428620664      0
#> 70   -0.7045150138      0
#> 71    0.6571256910      0
#> 72   -0.3194607603      0
#> 73    1.8219336893      0
#> 74    0.4107002306      0
#> 75    0.7812195894      0
#> 76    1.0091792047      0
#> 77   -0.5901683319      0
#> 78    0.0028199215      0
#> 79   -1.8777111660      0
#> 80   -1.6296292554      0
#> 81   -1.9888140679      0
#> 82    0.4575912020      0
#> 83   -1.1064969864      0
#> 84   -0.4978041343      0
#> 85    1.2875185927      0
#> 86    0.4754659587      0
#> 87    0.2749455219      0
#> 88    2.4571166211      0
#> 89   -0.6706464206      0
#> 90    0.1276203223      0
#> 91   -1.0677953719      0
#> 92    1.2040697201      0
#> 93    0.4313450162      0
#> 94   -0.1262102836      0
#> 95   -0.1647268293      0
#> 96   -0.0318272729      0
#> 97    2.3595512609      0
#> 98   -0.2684860015      0
#> 99    1.3009437593      0
#> 100   0.9852820259      0
#> 101  -0.0835640192      0
#> 102  -1.2164377791      0
#> 103  -0.2692301690      0
#> 104  -0.3002712485      0
#> 105  -0.9212031184      0
#> 106   0.6437750980      0
#> 107   0.2321909450      0
#> 108   1.1587526487      0
#> 109   0.6173450014      0
#> 110  -0.1166784964      0
#> 111   0.5253381038      0
#> 112   1.5867034529      0
#> 113   0.6542513218      0
#> 114   1.3062158853      0
#> 115  -1.5969262974      0
#> 116  -1.3055139706      0
#> 117   0.8144893998      0
#> 118  -0.8577890023      0
#> 119  -0.4048509530      0
#> 120  -0.2730589668      0
#> 121  -0.2166617527      0
#> 122  -0.0056927575      0
#> 123  -1.3666603575      0
#> 124  -0.1262130858      0
#> 125   0.1122069306      0
#> 126   0.7072320564      0
#> 127   0.6440442405      0
#> 128   1.7303745784      0
#> 129  -0.7402490152      0
#> 130  -0.1326191186      0
#> 131  -1.1393271080      0
#> 132  -0.6676330394      0
#> 133  -1.6465179743      0
#> 134  -1.2153929686      0
#> 135  -1.4369526794      0
#> 136  -0.0630656019      0
#> 137  -1.0081254795      0
#> 138  -0.6956207083      0
#> 139  -0.8283941820      0
#> 140   0.7815263382      0
#> 141   0.6180314893      0
#> 142  -0.6873924008      0
#> 143  -0.1627714417      0
#> 144   1.2355707328      0
#> 145   0.1269141332      0
#> 146   0.7767768039      0
#> 147  -1.4795202318      0
#> 148   0.7339297484      0
#> 149   0.4574590392      0
#> 150  -1.5526515590      0
#> 151  -0.3707278210      0
#> 152  -0.5264948158      0
#> 153   0.2964001728      0
#> 154   0.6418954921      0
#> 155  -2.1114636942      0
#> 156   0.2612168326      0
#> 157  -0.0386609490      0
#> 158  -0.1602734226      0
#> 159  -0.3082973728      0
#> 160  -0.6495608365      0
#> 161  -0.5622028538      0
#> 162   2.2988163442      0
#> 163   2.1002647420      0
#> 164  -1.2654415365      0
#> 165  -0.7199992818      0
#> 166  -1.5042633975      0
#> 167  -0.3574394173      0
#> 168   0.8758922988      0
#> 169   2.6509589053      0
#> 170  -0.6713853092      0
#> 171  -0.6107781591      0
#> 172  -0.1783968514      0
#> 173  -0.0004825942      0
#> 174  -0.2190650296      0
#> 175   1.3638968566      0
#> 176   0.3272376424      0
#> 177   0.8000569157      0
#> 178   0.5107239345      0
#> 179  -1.6443152465      0
#> 180   0.2861966052      0
#> 181  -0.3977799158      0
#> 182  -0.1309555645      0
#> 183  -0.9647897964      0
#> 184   0.3048685189      0
#> 185   1.5309465899      0
#> 186   0.0568299107      0
#> 187   0.0363426711      0
#> 188   0.1756965111      0
#> 189   1.0819388840      0
#> 190   0.9320865462      0
#> 191  -0.0055952804      0
#> 192  -1.3958036276      0
#> 193   0.1144861815      0
#> 194   0.2784014841      0
#> 195  -0.1585875593      0
#> 196  -0.3569583650      0
#> 197  -0.2835392236      0
#> 198  -1.1540726193      0
#> 199   0.0798895804      0
#> 200   0.8328864976      0
#> 201  -0.8625935634      0
#> 202  -0.7610208276      0
#> 203   0.4977762452      0
#> 204   0.9367607711      0
#> 205   0.3690389667      0
#> 206  -0.0816546499      0
#> 207   0.0958193142      0
#> 208  -0.6782193560      0
#> 209  -0.4381277866      0
#> 210  -0.8557539758      0
#> 211  -0.1629000877      0
#> 212  -0.4951986177      0
#> 213   0.3924227414      0
#> 214  -0.3134901628      0
#> 215  -0.4158645288      0
#> 216  -0.3138154368      0
#> 217   0.7356986273      0
#> 218  -0.5975611242      0
#> 219   0.1886264221      0
#> 220   0.2549180629      0
#> 221   1.9010392072      0
#> 222  -0.5858108626      0
#> 223   0.7638428889      0
#> 224  -0.4628165335      0
#> 225  -0.0353304442      0
#> 226  -2.3388457009      0
#> 227  -0.4475530874      0
#> 228   1.4776015467      0
#> 229  -0.6969378834      0
#> 230  -0.0788165334      0
#> 231  -0.1594302549      0
#> 232  -1.4381472736      0
#> 233   0.1715509449      0
#> 234  -0.8981097162      0
#> 235  -0.4415764041      0
#> 236   0.5515234111      0
#> 237   0.9913612113      0
#> 238   0.4742670651      0
#> 239  -1.6833801965      0
#> 240   0.3111760450      0
#> 241   0.4862677164      0
#> 242  -0.0682139161      0
#> 243   0.1420453850      0
#> 244  -0.5960635516      0
#> 245  -1.0550275967      0
#> 246  -0.7064960324      0
#> 247   0.8275400019      0
#> 248   0.5568721408      0
#> 249  -1.3818622342      0
#> 250  -0.2760796987      0
#> 251  -1.2273701659      0
#> 252  -0.5141035493      0
#> 253  -0.3636578274      0
#> 254   0.2125256071      0
#> 255   0.3293414338      0
#> 256   0.3442094763      0
#> 257   0.3309730053      0
#> 258   0.7061775619      0
#> 259  -0.1653990504      0
#> 260   0.7202563242      0
#> 261  -0.4519880068      0
#> 262   0.3181998732      0
#> 263   0.1009167372      0
#> 264  -0.4635501229      0
#> 265  -0.4759210660      0
#> 266  -1.6795250049      0
#> 267   0.0989189796      0
#> 268  -1.3316415629      0
#> 269  -0.0052139780      0
#> 270   0.0113734027      0
#> 271  -1.7685373627      0
#> 272  -1.4733537455      0
#> 273  -1.8088018789      0
#> 274  -0.4201764432      0
#> 275   1.1877898309      0
#> 276   0.9755765833      0
#> 277  -0.1366122913      0
#> 278   0.6886700980      0
#> 279   0.7110334068      0
#> 280  -0.5237093226      0
#> 281   1.4732776099      0
#> 282   0.7494836329      0
#> 283  -1.9443330366      0
#> 284   1.5494708789      0
#> 285   0.4037327074      0
#> 286  -0.4824154417      0
#> 287  -0.6353999556      0
#> 288   0.7586113344      0
#> 289   2.6366350272      0
#> 290  -1.1362645569      0
#> 291   1.0373777934      0
#> 292   1.1987630764      0
#> 293  -0.5499281860      0
#> 294   1.9257381815      0
#> 295   1.9059881095      0
#> 296   0.3159020741      0
#> 297  -0.0261527225      0
#> 298  -1.5578641422      0
#> 299  -0.3375772204      0
#> 300   0.7993271881      0
#> 301   0.4860249706      0
#> 302   2.1817670155      0
#> 303  -0.0194984587      0
#> 304   2.3879842172      0
#> 305  -0.8329249061      0
#> 306   0.8930136043      0
#> 307   1.2628715759      0
#> 308   1.0413210481      0
#> 309  -0.4030441148      0
#> 310   0.1797772355      0
#> 311  -0.4834282340      0
#> 312   0.6650497008      0
#> 313   1.5263094469      0
#> 314  -1.5325502797      0
#> 315  -0.5912920772      0
#> 316   0.9765475476      0
#> 317  -0.3882963771      0
#> 318  -0.4935360453      0
#> 319   1.1478103669      0
#> 320  -0.5620314698      0
#> 321  -1.3432073627      0
#> 322  -0.2759280342      0
#> 323   0.6275104519      0
#> 324  -0.2297916680      0
#> 325  -0.3169389232      0
#> 326  -2.3792893917      0
#> 327  -1.9821116804      0
#> 328  -1.7213427355      0
#> 329  -0.2905229459      0
#> 330  -0.5697377366      0
#> 331  -0.1077120305      0
#> 332  -0.6623764063      0
#> 333   2.1450269469      0
#> 334   0.9065819935      0
#> 335   1.5050818952      0
#> 336  -0.5454161742      0
#> 337  -1.4950095063      0
#> 338  -0.5514233350      0
#> 339  -0.4043095348      0
#> 340  -0.3719387096      0
#> 341   0.0676498391      0
#> 342  -0.3901700009      0
#> 343  -1.2350394187      0
#> 344  -0.2851174545      0
#> 345  -0.1041585310      0
#> 346  -0.3508475566      0
#> 347  -0.7489554883      0
#> 348  -1.0020564125      0
#> 349  -0.8803356315      0
#> 350  -0.8085143959      0
#> 351   0.9955839370      0
#> 352  -0.5097728000      0
#> 353  -2.0136878768      0
#> 354   1.6318081972      0
#> 355  -0.4875257121      0
#> 356  -0.9364591002      0
#> 357   0.1039022196      0
#> 358   1.5101615032      0
#> 359  -0.5381377082      0
#> 360  -0.8649798398      0
#> 361  -0.9563080016      0
#> 362   0.3807054599      0
#> 363  -0.6820769964      0
#> 364  -0.2705250859      0
#> 365   1.2918027680      0
#> 366  -1.0054541591      0
#> 367   0.5388090723      0
#> 368   0.8779535577      0
#> 369  -0.3229646042      0
#> 370  -0.7551098994      0
#> 371  -0.0955428564      0
#> 372  -1.4343439040      0
#> 373   0.0925221625      0
#> 374  -2.2020689716      0
#> 375  -0.6756174199      0
#> 376  -0.3386049322      0
#> 377  -0.8797045119      0
#> 378  -0.0825965431      0
#> 379   0.1143677411      0
#> 380   0.9865435391      0
#> 381   1.9341229904      0
#> 382   0.1478503627      0
#> 383  -0.7213347619      0
#> 384  -0.8108276743      0
#> 385  -0.1878111062      0
#> 386  -2.0109391357      0
#> 387   2.0063302384      0
#> 388   0.6709615548      0
#> 389  -1.1137648225      0
#> 390   0.6154133528      0
#> 391   0.2994641668      0
#> 392   1.3036041006      0
#> 393  -0.0122655517      0
#> 394  -0.7277054579      0
#> 395   0.8498568883      0
#> 396   0.1964646509      0
#> 397  -0.9552298363      0
#> 398   0.0628527297      0
#> 399  -0.4196708285      0
#> 400   0.2875441696      0
#> 401   0.8183718312      0
#> 402   0.2947765617      0
#> 403   0.7218350240      0
#> 404   1.4512296870      0
#> 405   1.5739421951      0
#> 406  -0.2733191143      0
#> 407  -1.3157754205      0
#> 408   1.2149369983      0
#> 409   0.9971054652      0
#> 410  -0.4678400863      0
#> 411  -2.2176118086      0
#> 412  -0.7333772115      0
#> 413  -0.4506827210      0
#> 414   0.4367402939      0
#> 415  -0.4135125782      0
#> 416  -0.2409657768      0
#> 417   0.5962146515      0
#> 418   0.8511800214      0
#> 419  -0.7471154872      0
#> 420  -1.6377432448      0
#> 421   0.5011638143      0
#> 422  -0.1547231005      0
#> 423  -1.3799021797      0
#> 424   0.7934058759      0
#> 425  -0.2733180397      0
#> 426  -0.8695352778      0
#> 427   1.2029584190      0
#> 428   0.3427521204      0
#> 429  -0.9517571853      0
#> 430  -1.0030036983      0
#> 431  -1.2731474458      0
#> 432   0.4255312306      0
#> 433  -1.5109899286      0
#> 434  -1.7361916305      0
#> 435   0.6934752580      0
#> 436   0.4631533082      0
#> 437  -0.9839211920      0
#> 438  -0.6590600898      0
#> 439   0.1234403047      0
#> 440  -1.1056418109      0
#> 441  -0.3795603925      0
#> 442   1.8708375032      0
#> 443  -1.0221724484      0
#> 444  -0.0085408137      0
#> 445   1.3721477894      0
#> 446  -1.9161463671      0
#> 447  -0.0469905634      0
#> 448  -0.4350100580      0
#> 449   0.7905707760      0
#> 450   1.0645791209      0
#> 451   1.6373354867      0
#> 452  -0.2038391729      0
#> 453  -0.1875295065      0
#> 454   1.3618340939      0
#> 455   0.6973979855      0
#> 456  -0.1892472351      0
#> 457  -0.2087592643      0
#> 458   1.5371629662      0
#> 459   0.4246308025      0
#> 460   0.6817072477      0
#> 461  -1.2578427328      0
#> 462  -0.8965161447      0
#> 463   0.4158084529      0
#> 464   0.1747716037      0
#> 465   0.2354308204      0
#> 466   0.4465207058      0
#> 467   2.0192769059      0
#> 468  -1.3038914846      0
#> 469  -1.0629148756      0
#> 470  -0.2007403369      0
#> 471   0.6381534986      0
#> 472  -1.8011374546      0
#> 473  -0.2758607052      0
#> 474   2.4130056314      0
#> 475   0.1220515088      0
#> 476  -1.0508424308      0
#> 477   0.3217477718      0
#> 478   0.7942457832      0
#> 479   1.9216151917      0
#> 480   0.0002636837      0
#> 481  -1.4355400054      0
#> 482  -0.2067910059      0
#> 483  -0.4690825789      0
#> 484  -1.1574779091      0
#> 485  -0.0397877532      0
#> 486   2.7514127502      0
#> 487   1.6412718754      0
#> 488   0.1839138425      0
#> 489   0.3012080055      0
#> 490   1.1237093653      0
#> 491  -0.2320645351      0
#> 492  -0.0040905482      0
#> 493  -1.3502765282      0
#> 494  -0.7020878533      0
#> 495  -0.0967660273      0
#> 496   0.7666391682      0
#> 497   0.6276618790      0
#> 498  -0.1660878667      0
#> 499   1.0312011710      0
#> 500   0.6486783051      0
#> 501  -0.6523164259      0
#> 502   0.6780580352      0
#> 503   0.8641909857      0
#> 504   0.9684539845      0
#> 505  -1.2591205994      0
#> 506   0.2653657128      0
#> 507   1.8315807975      0
#> 508   0.4615394883      0
#> 509  -1.5196840207      0
#> 510   0.1890618659      0
#> 511  -0.5802034543      0
#> 512   0.0532722578      0
#> 513   3.2751729149      0
#> 514  -0.3637828596      0
#> 515   1.0092006879      0
#> 516  -1.3629599771      0
#> 517   0.1676458219      0
#> 518  -1.5263050569      0
#> 519  -1.3108703328      0
#> 520   1.3609760935      0
#> 521   0.3718245464      0
#> 522   0.7960805073      0
#> 523  -0.5401806724      0
#> 524   0.3100616292      0
#> 525  -0.1972542069      0
#> 526   0.9379744079      0
#> 527  -0.3077883319      0
#> 528  -0.4784651841      0
#> 529  -0.5297833242      0
#> 530   1.4624059121      0
#> 531  -0.6948511725      0
#> 532   0.6722936748      0
#> 533   0.8036622417      0
#> 534   3.1106900932      0
#> 535  -1.4536768467      0
#> 536  -0.0848849711      0
#> 537   0.2115995554      0
#> 538   1.1396983757      0
#> 539  -0.5067194895      0
#> 540   0.2106387317      0
#> 541  -0.0995329851      0
#> 542   0.4865480395      0
#> 543  -0.3221966786      0
#> 544   0.6107851601      0
#> 545   0.5028998691      0
#> 546   0.3309407851      0
#> 547   1.6272397517      0
#> 548   0.3091532321      0
#> 549   0.3294020765      0
#> 550   0.3063532683      0
#> 551  -0.4619695964      0
#> 552   0.9857194354      0
#> 553  -0.0750565132      0
#> 554   0.6365545189      0
#> 555   0.2364740433      0
#> 556   2.2356145599      0
#> 557  -0.3160677160      0
#> 558  -0.7579822940      0
#> 559  -0.4941916023      0
#> 560  -0.7500579330      0
#> 561   0.7621639015      0
#> 562  -0.4933341274      0
#> 563   1.2931281907      0
#> 564  -0.1891217801      0
#> 565  -0.3665853321      0
#> 566   0.2585496405      0
#> 567   0.2050132314      0
#> 568  -1.5776277828      0
#> 569  -1.6671186381      0
#> 570   1.5674316678      0
#> 571   1.4601084218      0
#> 572   0.9408550824      0
#> 573   1.0533166697      0
#> 574  -2.1569907975      0
#> 575  -1.9508351495      0
#> 576   1.0739363412      0
#> 577  -1.0413328635      0
#> 578   0.2098821045      0
#> 579   0.7628571359      0
#> 580  -2.3811892024      0
#> 581   0.7797472644      0
#> 582  -0.0456252506      0
#> 583  -0.2374555013      0
#> 584  -0.3794025407      0
#> 585   0.1161710571      0
#> 586  -0.9611722368      0
#> 587  -0.4041832311      0
#> 588  -2.0772040399      0
#> 589   0.6978180399      0
#> 590  -1.4470830812      0
#> 591   1.7325030811      0
#> 592   0.2227710733      0
#> 593   0.5961600126      0
#> 594  -1.3476871356      0
#> 595   1.5685520548      0
#> 596  -0.5433688512      0
#> 597   0.0984033068      0
#> 598   0.0149695702      0
#> 599   1.0004355841      0
#> 600  -0.8695864921      0
#> 601  -0.3954915443      0
#> 602   0.5369909579      0
#> 603   1.4357960662      0
#> 604   0.4989987631      0
#> 605   0.4574926247      0
#> 606  -0.2575800004      0
#> 607  -1.1766570882      0
#> 608  -0.7888404812      0
#> 609   0.2968926368      0
#> 610   0.3211168876      0
#> 611  -0.9859374505      0
#> 612  -1.1155557821      0
#> 613   0.1452618029      0
#> 614  -0.7556896163      0
#> 615   0.8011737199      0
#> 616  -0.7434658897      0
#> 617   0.6159477255      0
#> 618  -1.0401035742      0
#> 619   0.5660187794      0
#> 620  -0.1998208112      0
#> 621  -1.9119344223      0
#> 622  -1.4447526587      0
#> 623   0.4918281630      0
#> 624  -0.0594783891      0
#> 625  -2.3185058734      0
#> 626  -0.3489077154      0
#> 627  -0.0046131326      0
#> 628   0.7213037552      0
#> 629  -0.5551799608      0
#> 630   1.6258036475      0
#> 631  -1.8998617901      0
#> 632   0.4855339680      0
#> 633  -0.0357088808      0
#> 634   0.2978886714      0
#> 635   1.8977770540      0
#> 636   0.4648397131      0
#> 637   0.0232657806      0
#> 638  -1.5681026480      0
#> 639   1.5199187812      0
#> 640  -0.0124790520      0
#> 641   0.7195949544      0
#> 642  -1.7146569280      0
#> 643   2.4151748308      0
#> 644  -0.2050656207      0
#> 645  -0.6718330359      0
#> 646  -0.4024027874      0
#> 647   0.8715638134      0
#> 648   0.8488081998      0
#> 649  -0.1462849867      0
#> 650   0.9607580542      0
#> 651  -0.5303532877      0
#> 652   0.0869777074      0
#> 653   0.8684725155      0
#> 654  -0.7675985649      0
#> 655   0.7863449713      0
#> 656   0.1814910309      0
#> 657  -0.8649002430      0
#> 658  -0.2534450669      0
#> 659  -1.0702731427      0
#> 660   0.0519578071      0
#> 661   0.8161684822      0
#> 662   0.0721483201      0
#> 663   1.3079113427      0
#> 664   0.0593771209      0
#> 665   0.9192880048      0
#> 666   0.3677878855      0
#> 667  -1.0591727437      0
#> 668  -0.1253341514      0
#> 669  -0.4928843030      0
#> 670   3.0044275530      0
#> 671  -0.5340725806      0
#> 672   0.0643007269      0
#> 673   2.1522830441      0
#> 674   0.4810901042      0
#> 675  -1.5260124771      0
#> 676   0.4476179177      0
#> 677   1.3411761054      0
#> 678  -1.6972924810      0
#> 679  -1.0213908401      0
#> 680   1.1436408981      0
#> 681  -0.2592292048      0
#> 682   1.7325868694      0
#> 683  -0.6970819720      0
#> 684  -0.8078277162      0
#> 685  -0.4691689410      0
#> 686   0.1437933369      0
#> 687   0.1863110869      0
#> 688  -1.4351840947      0
#> 689  -0.7883344743      0
#> 690   1.9497816638      0
#> 691  -0.1979044811      0
#> 692  -0.1820571263      0
#> 693   1.7802332994      0
#> 694   0.0211973698      0
#> 695  -0.1984472219      0
#> 696   1.7511638326      0
#> 697  -0.9867307773      0
#> 698   1.4226384830      0
#> 699  -1.3893449527      0
#> 700  -0.0308913548      0
#> 701   0.3672458574      0
#> 702   0.3431025136      0
#> 703   1.0947297313      0
#> 704   0.2989031691      0
#> 705   0.4866471943      0
#> 706  -0.9454176484      0
#> 707  -1.4129501213      0
#> 708  -0.9928010481      0
#> 709   0.9820249565      0
#> 710   0.3999847992      0
#> 711   0.8850231992      0
#> 712  -0.1212323081      0
#> 713   0.3280146697      0
#> 714  -1.7529464522      0
#> 715   1.6140258145      0
#> 716   0.1623814319      0
#> 717  -2.0573577529      0
#> 718   0.4086426314      0
#> 719   1.5860925449      0
#> 720  -0.3964482770      0
#> 721   1.3286362522      0
#> 722  -1.0496550147      0
#> 723  -1.2148443784      0
#> 724  -0.5016392664      0
#> 725  -0.7152061603      0
#> 726  -0.7295663456      0
#> 727  -1.8811799385      0
#> 728  -0.6739075266      0
#> 729  -0.3569587841      0
#> 730  -1.0623351824      0
#> 731   0.2106018512      0
#> 732  -0.7466824377      0
#> 733   0.3999182236      0
#> 734   0.0425240627      0
#> 735   0.5462222701      0
#> 736   0.6301855591      0
#> 737   0.9436898952      0
#> 738  -0.5739407542      0
#> 739  -0.9253942961      0
#> 740   0.9207829112      0
#> 741  -0.6890901119      0
#> 742   0.8553098249      0
#> 743  -0.7531140752      0
#> 744   1.9714035594      0
#> 745  -1.9763741706      0
#> 746  -2.5317103842      0
#> 747  -1.6810994720      0
#> 748  -1.5189778424      0
#> 749   0.6649706780      0
#> 750  -0.0075871475      0
#> 751   0.8334268074      0
#> 752  -0.4683061874      0
#> 753   1.4083663582      0
#> 754  -0.5939801961      0
#> 755   0.2075400429      0
#> 756   2.0827471815      0
#> 757   1.3318887351      0
#> 758  -0.2210559402      0
#> 759  -0.2739141768      0
#> 760  -0.4046694047      0
#> 761   1.2869707098      0
#> 762  -0.0930734534      0
#> 763  -0.5034310893      0
#> 764  -0.3601615093      0
#> 765   0.0383052079      0
#> 766   0.3745733141      0
#> 767  -1.5427702073      0
#> 768  -1.7332998219      0
#> 769  -0.4752226215      0
#> 770   0.1142421908      0
#> 771   0.5665194237      0
#> 772  -0.5161557639      0
#> 773  -1.5657101588      0
#> 774  -0.7763186734      0
#> 775   1.0836851268      0
#> 776  -0.2849441163      0
#> 777   0.7689965973      0
#> 778   0.5987512594      0
#> 779  -1.9314909894      0
#> 780   0.0636633842      0
#> 781   0.2779491661      0
#> 782   0.0166747394      0
#> 783   0.0303402427      0
#> 784   0.2157577430      0
#> 785  -1.0262525817      0
#> 786  -1.5105282215      0
#> 787   1.0182910389      0
#> 788  -2.1697161647      0
#> 789  -0.4733677941      0
#> 790   0.1348961807      0
#> 791  -0.1717252706      0
#> 792   1.3349778645      0
#> 793  -1.6118861489      0
#> 794  -0.6418904112      0
#> 795   0.3985412303      0
#> 796  -0.0907586579      0
#> 797   0.0802539427      0
#> 798   0.4746103447      0
#> 799   1.7513839907      0
#> 800   0.1885317505      0
#> 801   1.0614333074      0
#> 802   0.6377545320      0
#> 803  -0.5850770159      0
#> 804  -1.1228207274      0
#> 805   1.4495315667      0
#> 806   1.2021201919      0
#> 807  -1.8018036734      0
#> 808  -1.3128746673      0
#> 809  -1.4838150367      0
#> 810  -1.0289628726      0
#> 811   0.4863999760      0
#> 812  -0.1570898212      0
#> 813  -0.8147656231      0
#> 814   0.5648132977      0
#> 815  -0.6923663887      0
#> 816   1.2599736926      0
#> 817   0.6903393447      0
#> 818   0.8682978632      0
#> 819  -0.8034197848      0
#> 820   0.6385697927      0
#> 821  -0.1938458402      0
#> 822  -0.1470676049      0
#> 823  -0.5781898732      0
#> 824   1.2072283242      0
#> 825   0.4116682767      0
#> 826   1.6094869510      0
#> 827   2.1111055687      0
#> 828  -1.1170164907      0
#> 829   0.0716336528      0
#> 830  -0.7800505488      0
#> 831  -1.9572844147      0
#> 832   0.5301583368      0
#> 833  -0.9298946798      0
#> 834   2.3086766497      0
#> 835  -1.0672264309      0
#> 836   0.1099485049      0
#> 837   0.6300166922      0
#> 838  -0.8950927961      0
#> 839  -1.7509184274      0
#> 840   0.2020053353      0
#> 841  -0.0876674206      0
#> 842   1.4832121430      0
#> 843   0.6700166279      0
#> 844  -1.9864498504      0
#> 845  -0.3853422039      0
#> 846  -0.8482481190      0
#> 847  -0.1496998066      0
#> 848  -0.3903156501      0
#> 849   0.3627508014      0
#> 850  -0.2311278576      0
#> 851   1.9863870301      0
#> 852  -1.9445198347      0
#> 853   1.1299761186      0
#> 854   0.7280128789      0
#> 855  -1.1566024436      0
#> 856   0.7807465719      0
#> 857   0.7825147054      0
#> 858   0.2473424942      0
#> 859  -0.0694865581      0
#> 860   0.1702628235      0
#> 861  -1.9235733637      0
#> 862   0.4596301568      0
#> 863  -0.1558566702      0
#> 864  -0.6475548181      0
#> 865  -0.7321776913      0
#> 866  -0.3072367336      0
#> 867   1.2645049628      0
#> 868   0.3169878509      0
#> 869   0.7492251150      0
#> 870  -0.1761715405      0
#> 871  -0.4848548426      0
#> 872   1.1044735651      0
#> 873   2.0050602117      0
#> 874   1.0164981179      0
#> 875   0.6539399800      0
#> 876   0.8502442726      0
#> 877   0.7205114964      0
#> 878   0.3999454038      0
#> 879  -0.9856750647      0
#> 880   0.7560631918      0
#> 881  -1.2419868919      0
#> 882   0.5175570431      0
#> 883  -0.1028064613      0
#> 884   0.2111566088      0
#> 885   0.3029585050      0
#> 886  -0.9206307533      0
#> 887  -0.3079486063      0
#> 888   0.1432520114      0
#> 889  -0.8820492962      0
#> 890   0.0661101461      0
#> 891   0.5407309480      0
#> 892   0.4700139512      0
#> 893   0.4398569932      1
#> 894   0.4432369961      1
#> 895   1.2907894837      1
#> 896   0.7687435653      1
#> 897   1.6269190363      1
#> 898   1.7465086306      1
#> 899   0.5659183294      1
#> 900   1.0680921463      1
#> 901   0.8798112868      1
#> 902   1.6960179724      1
#> 903  -0.6334982212      1
#> 904  -0.1446814534      1
#> 905   0.3218242626      1
#> 906  -0.6931011297      1
#> 907  -0.1441121159      1
#> 908   0.3960753393      1
#> 909   0.4322982030      1
#> 910   1.4911268038      1
#> 911   0.9445619231      1
#> 912   1.4523478140      1
#> 913   0.8325255882      1
#> 914  -0.1241365560      1
#> 915   3.1809258803      1
#> 916   1.3232707713      1
#> 917   1.7185531612      1
#> 918   0.7582504340      1
#> 919   0.2487941770      1
#> 920  -1.0735795509      1
#> 921   1.0453683525      1
#> 922   2.4970100713      1
#> 923   0.1314221608      1
#> 924  -0.1530867379      1
#> 925   1.2118231603      1
#> 926   0.9634030809      1
#> 927   1.0836487654      1
#> 928   0.6416890610      1
#> 929   2.3404180319      1
#> 930   1.0520371160      1
#> 931   0.1021317530      1
#> 932   0.7471428508      1
#> 933   2.5469974809      1
#> 934  -0.4367340494      1
#> 935   1.9417255587      1
#> 936   0.3876263235      1
#> 937  -0.9939100240      1
#> 938   1.2213421614      1
#> 939  -0.5492053469      1
#> 940   0.5546834924      1
#> 941   2.1288548173      1
#> 942   0.2899431183      1
#> 943  -0.2808331320      1
#> 944   0.3332252897      1
#> 945   1.2715561846      1
#> 946  -0.8920148709      1
#> 947   0.2350481208      1
#> 948   0.4617366506      1
#> 949   1.8644431821      1
#> 950   1.3641149479      1
#> 951   1.8309564575      1
#> 952   1.6249097491      1
#> 953   1.3965166014      1
#> 954  -0.6233563348      1
#> 955   2.4975196720      1
#> 956   0.3242239469      1
#> 957   1.5181659249      1
#> 958   1.9310977932      1
#> 959  -0.9285893711      1
#> 960   0.8126452516      1
#> 961   1.8019402398      1
#> 962   1.0123135854      1
#> 963   0.7533608736      1
#> 964   0.4289765245      1
#> 965   0.1628644645      1
#> 966  -1.7318188189      1
#> 967   1.1473727702      1
#> 968   1.8410077898      1
#> 969   0.0449607331      1
#> 970   1.9559140461      1
#> 971   1.3620307710      1
#> 972   1.0139842787      1
#> 973   0.7546449286      1
#> 974   0.1430443794      1
#> 975   1.3440505637      1
#> 976   1.3742028026      1
#> 977   0.3800469646      1
#> 978   1.5969892677      1
#> 979   1.7077573089      1
#> 980   0.5339726117      1
#> 981   1.7917472346      1
#> 982  -0.1345294877      1
#> 983   1.5511812667      1
#> 984   2.7507155453      1
#> 985  -0.1271596378      1
#> 986   2.1286605348      1
#> 987   2.1904076120      1
#> 988  -0.7335477350      1
#> 989  -0.8050275722      1
#> 990  -0.1012044216      1
#> 991   1.4964154652      1
#> 992  -0.2848411534      1
#> 993  -0.0270838447      1
#> 994  -0.1476678703      1
#> 995   0.3143726250      1
#> 996   3.2347537686      1
#> 997   1.8547325834      1
#> 998   1.1894225590      1
#> 999  -0.5397655947      1
#> 1000  2.3843288390      1