Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     0.5154450553      0
#> 2    -0.6114136107      0
#> 3    -0.6073672804      0
#> 4     0.8253133956      0
#> 5     0.0025224243      0
#> 6    -1.5574022123      0
#> 7     2.0124911683      0
#> 8     2.0797105562      0
#> 9    -0.8349057486      0
#> 10   -1.1906584127      0
#> 11    0.8086490594      0
#> 12    1.1805463350      0
#> 13   -0.9696029666      0
#> 14   -0.6439360214      0
#> 15   -0.8160766451      0
#> 16   -0.9258225370      0
#> 17    0.8959729761      0
#> 18    1.0966611622      0
#> 19   -0.5084706852      0
#> 20   -0.3091438819      0
#> 21    1.1895680494      0
#> 22    0.6263585379      0
#> 23   -2.0855581927      0
#> 24   -1.1381600218      0
#> 25    1.5247404245      0
#> 26   -0.2214364035      0
#> 27   -0.0926863526      0
#> 28   -0.1428468231      0
#> 29   -0.4618979835      0
#> 30   -1.3728394184      0
#> 31   -0.5919141411      0
#> 32    0.0708043292      0
#> 33    0.4447090743      0
#> 34   -1.1595108668      0
#> 35   -0.7188155319      0
#> 36   -1.3934686042      0
#> 37   -0.7071042495      0
#> 38    1.0089404041      0
#> 39   -0.1999105126      0
#> 40    1.0897675307      0
#> 41   -0.1069013712      0
#> 42   -2.0940701505      0
#> 43    1.0965775094      0
#> 44   -1.2093566501      0
#> 45    0.0104917995      0
#> 46    2.4469689722      0
#> 47    0.4370375812      0
#> 48   -1.7175100587      0
#> 49    0.8665356452      0
#> 50    0.7765330797      0
#> 51    1.1132283671      0
#> 52   -1.1906607875      0
#> 53   -1.2666008687      0
#> 54    0.2722801871      0
#> 55    0.3499334747      0
#> 56    1.5579547305      0
#> 57    1.0345506269      0
#> 58   -1.0234297366      0
#> 59    0.8774200353      0
#> 60   -0.3558595149      0
#> 61    1.0483455696      0
#> 62    0.7694962869      0
#> 63   -0.4165501484      0
#> 64    0.6927151352      0
#> 65    1.6790246287      0
#> 66   -0.8119753447      0
#> 67    0.8617049008      0
#> 68    0.8075893171      0
#> 69   -0.7161368097      0
#> 70   -0.0305805784      0
#> 71   -1.3601611918      0
#> 72   -0.1181113924      0
#> 73    0.9758625936      0
#> 74   -0.7390120589      0
#> 75   -0.7336835576      0
#> 76   -0.9104734436      0
#> 77   -0.5483143193      0
#> 78   -0.2042192517      0
#> 79    0.4187775646      0
#> 80    1.8750854379      0
#> 81   -0.5666765224      0
#> 82   -1.2819622967      0
#> 83   -0.6794264369      0
#> 84   -0.2477511895      0
#> 85   -1.7892672237      0
#> 86    0.8603777442      0
#> 87    0.6495421807      0
#> 88    1.8371537364      0
#> 89   -2.1393093617      0
#> 90   -1.5118122456      0
#> 91   -1.6241807556      0
#> 92   -1.8514961520      0
#> 93    0.2537966028      0
#> 94    0.7344168515      0
#> 95   -0.6541635980      0
#> 96   -0.8275223118      0
#> 97    0.3678112042      0
#> 98    0.9670823227      0
#> 99    1.0151330250      0
#> 100   0.1118410850      0
#> 101  -0.4441223109      0
#> 102   0.5351319998      0
#> 103  -0.7559408388      0
#> 104  -0.4361326522      0
#> 105   0.0640686328      0
#> 106  -1.0884460681      0
#> 107   1.7807359593      0
#> 108   0.2465077223      0
#> 109  -1.4629949431      0
#> 110   0.4052739428      0
#> 111  -2.5296054631      0
#> 112   0.8703004756      0
#> 113   0.2672422425      0
#> 114   0.1768616880      0
#> 115   0.0129249127      0
#> 116  -0.4218410540      0
#> 117   1.1219602572      0
#> 118   0.2156343194      0
#> 119   0.5283610205      0
#> 120   0.3862180336      0
#> 121  -0.3866399040      0
#> 122  -0.8235829240      0
#> 123   0.2930336993      0
#> 124  -0.0532482363      0
#> 125   1.5208052269      0
#> 126  -0.0385033623      0
#> 127   0.0319099247      0
#> 128  -0.8682631967      0
#> 129  -1.3446169858      0
#> 130   0.0029843108      0
#> 131   2.1247231108      0
#> 132   0.7418235455      0
#> 133   0.6835745616      0
#> 134  -1.1699925992      0
#> 135  -0.1636623556      0
#> 136   1.0718308494      0
#> 137   1.0940449886      0
#> 138   0.6622787319      0
#> 139   1.0816871967      0
#> 140  -1.0922445995      0
#> 141  -0.5696029700      0
#> 142  -0.2512180731      0
#> 143   0.9435002133      0
#> 144   0.5168745868      0
#> 145  -0.2550616820      0
#> 146   1.2963674675      0
#> 147   0.3722430008      0
#> 148   0.2242587133      0
#> 149   0.7101627517      0
#> 150   1.3719392590      0
#> 151   1.7199285486      0
#> 152   0.4731975812      0
#> 153   0.8468459515      0
#> 154   0.2802274659      0
#> 155   1.3799592100      0
#> 156   0.4977576452      0
#> 157  -0.8485911929      0
#> 158   0.5195887829      0
#> 159   0.5141882325      0
#> 160  -1.3693962286      0
#> 161   0.3734696145      0
#> 162   0.0296079328      0
#> 163  -0.4163822821      0
#> 164  -1.8582583331      0
#> 165   0.4514545262      0
#> 166  -1.9470973122      0
#> 167   0.4976375166      0
#> 168   0.0436020640      0
#> 169  -1.7363346582      0
#> 170  -0.3802979558      0
#> 171   1.8445099691      0
#> 172   0.1539589211      0
#> 173  -1.2521417721      0
#> 174  -0.0813408916      0
#> 175   0.3661765324      0
#> 176  -0.7193380374      0
#> 177  -0.4300579012      0
#> 178   0.3199431971      0
#> 179   0.0188212661      0
#> 180  -0.1140890213      0
#> 181  -0.2878964882      0
#> 182  -1.5632163627      0
#> 183  -0.1577727205      0
#> 184  -0.9139336967      0
#> 185  -0.1974296749      0
#> 186   0.2929410553      0
#> 187  -1.6821268119      0
#> 188   0.9343253336      0
#> 189  -0.0719222737      0
#> 190  -0.9007733226      0
#> 191  -1.8419120382      0
#> 192   0.6582287125      0
#> 193   1.6336604673      0
#> 194   1.0043373930      0
#> 195  -0.9536934632      0
#> 196   0.4161582357      0
#> 197  -1.7170709100      0
#> 198   0.4294559918      0
#> 199  -0.9970757459      0
#> 200  -0.9548819459      0
#> 201  -1.8248118927      0
#> 202  -0.6070927490      0
#> 203  -0.7256257266      0
#> 204  -2.0037668606      0
#> 205  -0.3699995748      0
#> 206  -0.0655524161      0
#> 207   0.9984783974      0
#> 208  -0.1373808224      0
#> 209   0.9963223523      0
#> 210   0.3503488934      0
#> 211  -2.4559699184      0
#> 212  -0.5529850024      0
#> 213   0.5404490830      0
#> 214  -0.5825703014      0
#> 215   1.2492013427      0
#> 216  -0.0281584269      0
#> 217  -0.4482827759      0
#> 218  -0.4262406506      0
#> 219   0.5770273507      0
#> 220   0.8425318684      0
#> 221   1.3121130765      0
#> 222  -2.0315325189      0
#> 223   1.1701372609      0
#> 224  -1.6001891724      0
#> 225  -0.2570714453      0
#> 226   0.9570525880      0
#> 227  -0.1954501058      0
#> 228  -1.3382867328      0
#> 229  -0.4277749416      0
#> 230  -1.6672470114      0
#> 231   1.4526015586      0
#> 232  -0.3881933030      0
#> 233   1.1135204340      0
#> 234   0.6098659494      0
#> 235   0.2297504472      0
#> 236  -0.3272724685      0
#> 237   1.0426148117      0
#> 238  -0.1494580711      0
#> 239   1.9268258900      0
#> 240   1.4797456807      0
#> 241   0.0254691639      0
#> 242  -0.4392279672      0
#> 243  -1.5402181934      0
#> 244   0.2168781607      0
#> 245   0.7553356728      0
#> 246   0.3061235848      0
#> 247   0.0846152224      0
#> 248   0.5065697075      0
#> 249   0.2252342064      0
#> 250   0.0634639287      0
#> 251  -0.3091009488      0
#> 252   1.0374341585      0
#> 253  -0.2306792815      0
#> 254   1.3446426680      0
#> 255  -1.8707061915      0
#> 256   0.7611403035      0
#> 257  -0.9506863298      0
#> 258  -1.5025473288      0
#> 259   0.3210143026      0
#> 260   1.1152679162      0
#> 261  -1.3040103849      0
#> 262  -1.0612355481      0
#> 263  -2.0911357499      0
#> 264  -0.2756441421      0
#> 265   0.1395756685      0
#> 266   0.6720013032      0
#> 267   1.3709574352      0
#> 268   2.1243879787      0
#> 269  -2.0622580716      0
#> 270   1.5559712188      0
#> 271  -2.4984062439      0
#> 272   0.2451430615      0
#> 273  -0.9458450353      0
#> 274   1.3812056686      0
#> 275  -0.1331271609      0
#> 276  -1.9945941202      0
#> 277   0.3716545124      0
#> 278  -0.0941243598      0
#> 279  -1.8553835223      0
#> 280  -1.0643809707      0
#> 281  -0.6803176248      0
#> 282   1.0386163607      0
#> 283  -0.3698582168      0
#> 284   0.3714073584      0
#> 285  -1.5196979444      0
#> 286  -1.2283262141      0
#> 287  -0.5717191153      0
#> 288  -0.4025663812      0
#> 289  -1.8063210828      0
#> 290  -0.1228581700      0
#> 291  -0.2298030988      0
#> 292   0.1715443849      0
#> 293   0.8090407630      0
#> 294  -0.1834019854      0
#> 295  -0.1822265495      0
#> 296  -1.2346682218      0
#> 297   3.3249557629      0
#> 298   0.1806993158      0
#> 299   1.3777708619      0
#> 300  -0.3500905541      0
#> 301   0.1568412083      0
#> 302  -0.8458253462      0
#> 303  -0.5362172392      0
#> 304  -0.7014575699      0
#> 305  -0.3654849676      0
#> 306   0.6190162544      0
#> 307   0.6332958983      0
#> 308  -0.5541714233      0
#> 309   0.4169396816      0
#> 310  -1.3742893854      0
#> 311   1.6757463640      0
#> 312  -0.3175363014      0
#> 313   1.4749275172      0
#> 314   1.4194711145      0
#> 315  -0.8487268406      0
#> 316  -1.1311153750      0
#> 317   0.3726458650      0
#> 318  -0.9217477644      0
#> 319   1.0726036760      0
#> 320   0.6206469537      0
#> 321   1.7062447558      0
#> 322   1.1167780281      0
#> 323  -0.2140724557      0
#> 324  -0.3285858472      0
#> 325   0.4810176996      0
#> 326  -0.5974373680      0
#> 327  -1.2757022548      0
#> 328  -0.5484876677      0
#> 329  -0.0271676873      0
#> 330  -1.3659301138      0
#> 331   1.5613746490      0
#> 332   1.4973016510      0
#> 333   0.0067913846      0
#> 334   0.8402697435      0
#> 335   0.6694606562      0
#> 336   1.2519377994      0
#> 337   1.9977614699      0
#> 338   0.5933544553      0
#> 339  -0.5461506077      0
#> 340  -2.0732720655      0
#> 341  -0.2247124398      0
#> 342   0.0880837246      0
#> 343  -0.6653874272      0
#> 344   1.3305499956      0
#> 345   0.1930039267      0
#> 346  -1.5609327491      0
#> 347   0.2777157681      0
#> 348  -0.2605815557      0
#> 349  -0.5139681631      0
#> 350   1.0067807069      0
#> 351   1.1492631432      0
#> 352  -0.6914155008      0
#> 353   0.7608664175      0
#> 354  -0.0802647108      0
#> 355   1.4848589121      0
#> 356   1.6310479755      0
#> 357  -0.4611596634      0
#> 358  -0.0516894516      0
#> 359   0.0470607139      0
#> 360   0.2114203088      0
#> 361  -0.9063812961      0
#> 362   0.3871046199      0
#> 363   0.2405939060      0
#> 364  -1.0876054545      0
#> 365   1.1530863503      0
#> 366  -1.2324440250      0
#> 367  -0.2955979837      0
#> 368   0.3310864703      0
#> 369  -0.8998245708      0
#> 370  -1.4495387518      0
#> 371   0.9313495263      0
#> 372  -0.1511371517      0
#> 373  -0.7711129159      0
#> 374  -1.1136533267      0
#> 375   0.9821249086      0
#> 376   0.4047525566      0
#> 377   0.4300888741      0
#> 378   0.4620685685      0
#> 379   0.1099983220      0
#> 380  -0.5925820691      0
#> 381  -2.1928302854      0
#> 382   0.0201736335      0
#> 383   1.2583309016      0
#> 384   0.7336779330      0
#> 385   0.6094370190      0
#> 386  -1.3244123502      0
#> 387  -0.7883079034      0
#> 388  -1.7024087260      0
#> 389   1.0896274577      0
#> 390   0.9504571105      0
#> 391  -0.5731514987      0
#> 392  -1.4196281469      0
#> 393  -1.3499392559      0
#> 394  -0.0614314127      0
#> 395  -0.9523748946      0
#> 396  -0.7940070681      0
#> 397  -1.8442048306      0
#> 398  -0.7111795550      0
#> 399   1.4880159152      0
#> 400   0.4307753256      0
#> 401   0.2639187635      0
#> 402   1.7470800245      0
#> 403   0.4005384282      0
#> 404  -0.8178406250      0
#> 405  -1.0255390137      0
#> 406  -0.0677097797      0
#> 407  -0.7785392082      0
#> 408  -0.0947740787      0
#> 409   0.6045471837      0
#> 410  -0.4869517051      0
#> 411  -1.1736083672      0
#> 412  -0.6811387124      0
#> 413  -0.4369373235      0
#> 414   1.6170883802      0
#> 415   0.1725560189      0
#> 416  -0.3855344509      0
#> 417  -1.7950519025      0
#> 418  -0.6034478369      0
#> 419  -0.4670463783      0
#> 420   0.4491609265      0
#> 421  -0.6699157932      0
#> 422   1.4759590564      0
#> 423   0.1201517377      0
#> 424   0.7989381398      0
#> 425   2.3481444037      0
#> 426   0.6770585043      0
#> 427   0.4290274986      0
#> 428   1.1707037423      0
#> 429  -1.4061880060      0
#> 430   0.8684373093      0
#> 431   0.1940437837      0
#> 432   0.0724908967      0
#> 433   1.1504705343      0
#> 434  -0.3635542157      0
#> 435   0.2866751912      0
#> 436   1.0361623162      0
#> 437   0.6623126547      0
#> 438   0.3940836974      0
#> 439  -0.1268026506      0
#> 440  -0.8068148548      0
#> 441   0.1810211065      0
#> 442   0.0964943512      0
#> 443   2.5469383756      0
#> 444  -0.4842006991      0
#> 445   0.3780024201      0
#> 446   0.7158581123      0
#> 447   0.9528135486      0
#> 448   1.3411835778      0
#> 449   0.6092707813      0
#> 450   0.4493459833      0
#> 451   1.3725625348      0
#> 452  -0.5147842193      0
#> 453   2.2474278576      0
#> 454  -0.4913721624      0
#> 455  -1.1532711275      0
#> 456  -0.1997355767      0
#> 457  -1.7397505541      0
#> 458   0.4460976376      0
#> 459   0.0409365252      0
#> 460  -0.1865493271      0
#> 461   1.0716235976      0
#> 462   1.6280740093      0
#> 463  -0.2049757672      0
#> 464   1.1883565993      0
#> 465   1.1593038037      0
#> 466  -1.2060452196      0
#> 467   0.2218095737      0
#> 468   1.3479896398      0
#> 469  -1.2532008505      0
#> 470   2.0740112753      0
#> 471   2.4511427939      0
#> 472   0.3662286061      0
#> 473   0.1124822668      0
#> 474  -1.6521455310      0
#> 475   1.9894365253      0
#> 476   1.0452844449      0
#> 477  -0.9743492922      0
#> 478  -0.3455316971      0
#> 479  -1.4396247404      0
#> 480   0.0022974026      0
#> 481   0.4722575048      0
#> 482  -0.7175212452      0
#> 483   2.7315744118      0
#> 484   0.0001293511      0
#> 485   0.7310382083      0
#> 486   0.4940718893      0
#> 487  -0.8856752831      0
#> 488   0.3767417300      0
#> 489   0.3532775806      0
#> 490  -0.1562015882      0
#> 491  -2.4083020823      0
#> 492  -0.4696833593      0
#> 493   0.4499363894      0
#> 494  -1.2908646371      0
#> 495   0.4618282424      0
#> 496  -1.5244278024      0
#> 497  -0.8106226747      0
#> 498   0.1554643907      0
#> 499  -0.1580550190      0
#> 500   1.4480670163      0
#> 501   0.9127028603      0
#> 502  -0.1764865792      0
#> 503   0.9998303698      0
#> 504   1.6003847215      0
#> 505  -0.5495950113      0
#> 506   0.5064810136      0
#> 507   0.8245323161      0
#> 508  -1.0175771876      0
#> 509   1.6541930704      0
#> 510   0.7082808310      0
#> 511   0.1425237801      0
#> 512  -0.6181304056      0
#> 513   1.0212063944      0
#> 514  -0.4868223029      0
#> 515   1.9592632003      0
#> 516   0.4209304262      0
#> 517   2.0624797099      0
#> 518   0.7049524933      0
#> 519  -2.1983737546      0
#> 520  -0.6108673161      0
#> 521   0.8816969327      0
#> 522  -0.0058319178      0
#> 523  -0.2143654321      0
#> 524  -0.8261983327      0
#> 525   0.2761644053      0
#> 526   0.9165602936      0
#> 527   0.2886526278      0
#> 528  -0.2950344109      0
#> 529  -1.7357938919      0
#> 530  -0.7417317124      0
#> 531  -0.0745349509      0
#> 532   0.0547502607      0
#> 533   0.9601621992      0
#> 534   0.1154052675      0
#> 535  -1.2213699934      0
#> 536  -0.4555411857      0
#> 537   0.1610860840      0
#> 538  -1.4080831285      0
#> 539   0.1051593775      0
#> 540   2.3462947810      0
#> 541  -0.6175512042      0
#> 542   0.3108430650      0
#> 543  -0.7952817638      0
#> 544  -0.6436497080      0
#> 545   0.0899572119      0
#> 546   0.2352109794      0
#> 547  -0.8148555582      0
#> 548  -0.0594898279      0
#> 549   1.3798650481      0
#> 550  -1.7628648846      0
#> 551  -1.4429351891      0
#> 552   0.9864678237      0
#> 553   1.3854574137      0
#> 554   0.7650189644      0
#> 555   0.4662530402      0
#> 556   0.9181958465      0
#> 557   0.5122468338      0
#> 558   0.5274052040      0
#> 559   0.0480203729      0
#> 560   0.8800400358      0
#> 561  -0.1636817933      0
#> 562   1.1563603972      0
#> 563   0.0481390644      0
#> 564  -0.1902772773      0
#> 565  -0.9830139045      0
#> 566   0.5405473986      0
#> 567  -0.1674815904      0
#> 568  -0.2247147205      0
#> 569   1.8457373514      0
#> 570  -0.8671522868      0
#> 571   0.7848399793      0
#> 572  -1.3615606228      0
#> 573  -2.0550840494      0
#> 574  -0.8670014453      0
#> 575  -0.4521135048      0
#> 576  -0.6868738801      0
#> 577   1.6504884469      0
#> 578  -0.4624636284      0
#> 579  -1.2894628866      0
#> 580  -0.3563744226      0
#> 581  -0.0744564812      0
#> 582   0.4997684961      0
#> 583  -0.5784892293      0
#> 584  -0.1887490912      0
#> 585   0.5895408020      0
#> 586  -1.6672014502      0
#> 587   0.6928637870      0
#> 588  -0.8393300633      0
#> 589  -1.0076908718      0
#> 590  -0.4998782503      0
#> 591   0.3990046545      0
#> 592  -1.4060039718      0
#> 593  -0.5226623908      0
#> 594   0.8841647838      0
#> 595  -0.9157646069      0
#> 596   1.0954565516      0
#> 597  -1.9054115685      0
#> 598  -0.1016364036      0
#> 599   1.3588947937      0
#> 600   0.2163145840      0
#> 601   2.3679977079      0
#> 602  -1.0911239613      0
#> 603  -0.2699304665      0
#> 604   1.7355186178      0
#> 605  -0.1278282705      0
#> 606   0.2047950358      0
#> 607   1.0938308264      0
#> 608   1.3189502591      0
#> 609   0.4277158637      0
#> 610  -0.0269861920      0
#> 611  -0.3143145178      0
#> 612   0.7148046348      0
#> 613   0.3300797462      0
#> 614  -1.1557965918      0
#> 615   0.0869496434      0
#> 616   1.4016438595      0
#> 617   0.2500497809      0
#> 618   0.7129928169      0
#> 619   2.3310079504      0
#> 620  -0.6948276171      0
#> 621  -0.6418999406      0
#> 622   1.9247769539      0
#> 623   0.7458339190      0
#> 624   1.2639094970      0
#> 625  -0.3622353026      0
#> 626  -0.5436863692      0
#> 627  -0.7307757879      0
#> 628   0.1031788528      0
#> 629   0.5719145144      0
#> 630   1.6865243182      0
#> 631  -1.9968155569      0
#> 632  -0.4764571744      0
#> 633   1.6011250935      0
#> 634   0.0848179591      0
#> 635   0.6572747544      0
#> 636  -0.0716278671      0
#> 637   1.1801576359      0
#> 638  -2.1766593651      0
#> 639  -0.2906147180      0
#> 640  -0.9546791543      0
#> 641   2.3611199229      0
#> 642  -0.6181589376      0
#> 643   1.6646085703      0
#> 644  -1.2053010735      0
#> 645  -0.5786224451      0
#> 646   0.0634922772      0
#> 647   1.4790535816      0
#> 648  -0.3723558722      0
#> 649   1.7256253894      0
#> 650  -1.0335524428      0
#> 651  -0.9342727349      0
#> 652  -0.4918451461      0
#> 653  -0.1680320646      0
#> 654   0.6322980727      0
#> 655   0.5112627170      0
#> 656   0.1369959256      0
#> 657   0.6215672684      0
#> 658  -0.7330319096      0
#> 659  -0.7168489644      0
#> 660  -0.4308128748      0
#> 661  -0.7849648874      0
#> 662  -0.4744225130      0
#> 663   0.3420042748      0
#> 664   0.0322794620      0
#> 665  -0.4619637590      0
#> 666  -0.3627213159      0
#> 667  -0.0410399886      0
#> 668  -0.1213147520      0
#> 669  -0.5277352543      0
#> 670  -1.0522329001      0
#> 671   0.0469630040      0
#> 672   1.6185503481      0
#> 673   1.0482913012      0
#> 674   0.1772747206      0
#> 675  -0.8686405149      0
#> 676   1.7194804820      0
#> 677  -1.9476821400      0
#> 678  -1.7856259501      0
#> 679   0.3289878909      0
#> 680  -0.8373478822      0
#> 681   2.7974612141      0
#> 682  -0.9003315641      0
#> 683  -0.2010071150      0
#> 684  -1.1538181928      0
#> 685   0.1521567654      0
#> 686  -0.5987938014      0
#> 687  -0.6913385029      0
#> 688  -3.0496341165      0
#> 689   0.4336003965      0
#> 690   0.6275206150      0
#> 691  -1.8341466948      0
#> 692   0.5377598500      0
#> 693  -2.4121510126      0
#> 694   0.6473331118      0
#> 695  -2.2946334679      0
#> 696  -0.9448214694      0
#> 697   1.6242032311      0
#> 698  -1.6720678684      0
#> 699   0.9824676440      0
#> 700  -0.0112339374      0
#> 701  -0.6060892174      0
#> 702   0.1888804385      0
#> 703  -1.9854877372      0
#> 704   0.5026699721      0
#> 705   0.4864911163      0
#> 706   0.4558018189      0
#> 707   0.2125488959      0
#> 708   0.6269531976      0
#> 709   0.3664662304      0
#> 710   0.1706735769      0
#> 711   1.1595263973      0
#> 712   1.4902260682      0
#> 713  -1.4327978889      0
#> 714  -0.7694424077      0
#> 715   0.4578281134      0
#> 716   1.3422469001      0
#> 717  -0.0205238948      0
#> 718   1.0683378930      0
#> 719   0.4474490195      0
#> 720   0.0484711134      0
#> 721  -1.4154367439      0
#> 722  -1.9723867017      0
#> 723  -0.1772601634      0
#> 724   0.7302447916      0
#> 725  -0.9258672009      0
#> 726   0.9526343005      0
#> 727  -0.0639376217      0
#> 728   1.0270011495      0
#> 729  -2.3595318858      0
#> 730   0.3811481005      0
#> 731   1.3728620412      0
#> 732  -0.0715146778      0
#> 733   0.8178254617      0
#> 734   1.1600211384      0
#> 735   2.1042821458      0
#> 736  -2.3214037845      0
#> 737  -0.2756923868      0
#> 738   0.2357901780      0
#> 739   0.1923940563      0
#> 740   0.6085128790      0
#> 741  -0.0298080577      0
#> 742  -1.3222065670      0
#> 743   0.3826432811      0
#> 744  -0.5697204436      0
#> 745   0.1308415903      0
#> 746   1.1440597683      0
#> 747  -1.1162773433      0
#> 748  -0.8841523023      0
#> 749   1.6334826053      0
#> 750  -0.7466830446      0
#> 751  -1.6775803845      0
#> 752  -0.4043249274      0
#> 753  -2.6575838425      0
#> 754  -0.0716686678      0
#> 755  -2.1633525558      0
#> 756  -0.5994452934      0
#> 757   0.4697365947      0
#> 758   0.7363499719      0
#> 759   1.1596911787      0
#> 760   1.3401861554      0
#> 761  -0.3632189118      0
#> 762   1.7170532486      0
#> 763   0.2831981398      0
#> 764  -0.1357629879      0
#> 765  -0.0444372241      0
#> 766  -0.1037634512      0
#> 767   0.1032380190      0
#> 768  -0.0761557919      0
#> 769   1.4456519380      0
#> 770  -0.3931657606      0
#> 771   0.2214473133      0
#> 772   0.0365702294      0
#> 773   1.4923757445      0
#> 774   0.0207774016      0
#> 775  -0.5213504272      0
#> 776  -0.7079523264      0
#> 777  -1.2300383003      0
#> 778   0.4181038408      0
#> 779   0.0143234240      0
#> 780  -1.4865230048      0
#> 781  -0.5852570844      0
#> 782   2.0375117411      0
#> 783  -1.7827069529      0
#> 784   0.7812723911      0
#> 785   0.3249786792      0
#> 786  -0.1124689852      0
#> 787   0.7415912080      0
#> 788   0.8587730875      0
#> 789   0.2491318760      0
#> 790   1.5574265244      0
#> 791   0.3941902843      0
#> 792  -0.0974400628      0
#> 793  -0.4702930521      0
#> 794   0.0056350245      0
#> 795  -0.6997513970      0
#> 796  -0.8609662443      0
#> 797  -0.5355507375      0
#> 798   0.5973588237      0
#> 799  -0.2714581435      0
#> 800   0.3979116887      0
#> 801   1.3579081518      0
#> 802  -1.6642058056      0
#> 803   1.3057977787      0
#> 804  -0.2886564232      0
#> 805   0.6540217110      0
#> 806  -2.1539323296      0
#> 807  -1.5621879192      0
#> 808  -0.2393530856      0
#> 809  -0.6337476646      0
#> 810   0.9828827188      0
#> 811  -0.0675799696      0
#> 812   0.3266544823      0
#> 813  -0.5744886394      0
#> 814  -0.1596326218      0
#> 815  -0.3203008882      0
#> 816   1.5105897277      0
#> 817  -0.5264564116      0
#> 818   1.2492179356      0
#> 819  -0.5279115388      0
#> 820   0.7342703199      0
#> 821  -0.1300288960      0
#> 822  -0.0776513778      0
#> 823   1.1654503430      0
#> 824  -0.5294849682      0
#> 825   1.6819033056      0
#> 826  -0.1905030159      0
#> 827  -0.3105480132      0
#> 828   1.3744759062      0
#> 829  -0.1497112610      0
#> 830  -0.1210839186      0
#> 831   0.8175594156      0
#> 832  -1.1486426152      0
#> 833   0.3980177303      0
#> 834   1.0433772048      0
#> 835  -0.0402123620      0
#> 836  -0.1308760949      0
#> 837   0.0651241550      0
#> 838  -1.1756458247      0
#> 839  -0.9986299088      0
#> 840  -0.5587581023      0
#> 841   1.6423216198      0
#> 842   3.1800025550      0
#> 843  -1.2187210209      0
#> 844  -0.8545683625      0
#> 845  -1.6405683567      0
#> 846  -0.1391190562      0
#> 847  -0.1825645857      0
#> 848   1.2317914431      0
#> 849   0.5485925225      0
#> 850   1.7267793420      0
#> 851  -0.7001386269      0
#> 852   0.4262743833      0
#> 853   0.7894617166      0
#> 854  -0.8599534160      0
#> 855  -0.5431761370      0
#> 856   0.0152997498      0
#> 857   0.2140969854      0
#> 858   1.5087783215      0
#> 859   1.4530752816      0
#> 860  -0.1063079762      0
#> 861  -2.5171446872      0
#> 862  -1.1683595434      0
#> 863   0.7903298837      0
#> 864   1.5252654666      0
#> 865  -0.3853517964      0
#> 866   0.4345760792      0
#> 867   1.0468027875      0
#> 868  -1.8789291604      0
#> 869   0.3733899930      0
#> 870   0.0239445667      0
#> 871   0.2435715138      0
#> 872  -0.9238372675      0
#> 873  -1.5905798515      0
#> 874   1.2586681328      0
#> 875   0.5185551142      0
#> 876   0.4584822707      0
#> 877  -0.4611614286      0
#> 878   0.0941823547      0
#> 879   1.5423704137      0
#> 880   0.0027564403      0
#> 881  -0.1293519155      0
#> 882   0.5810769054      0
#> 883  -0.5611229838      0
#> 884   0.3252395750      0
#> 885  -0.8990354514      0
#> 886  -0.1102157194      0
#> 887   0.2978111472      0
#> 888  -0.6233545377      0
#> 889   0.5854202093      0
#> 890  -0.6027413492      0
#> 891   0.4028073656      0
#> 892   1.8392593847      0
#> 893   0.0916322790      0
#> 894   0.8646392670      0
#> 895   0.0295762502      0
#> 896   0.1715658448      0
#> 897   0.2446890448      0
#> 898   0.5800822162      1
#> 899   0.2469474886      1
#> 900   0.1154894969      1
#> 901   0.6912831576      1
#> 902   0.4129684959      1
#> 903   0.7996296234      1
#> 904   2.3079359338      1
#> 905   0.1748554767      1
#> 906   1.1565289918      1
#> 907   0.9239256908      1
#> 908   2.4646174231      1
#> 909  -0.2658789619      1
#> 910   2.3930016590      1
#> 911   0.5927814625      1
#> 912   0.9670627636      1
#> 913   1.4385377892      1
#> 914   0.0001384197      1
#> 915   1.6576421541      1
#> 916   1.7660819384      1
#> 917   2.4560579416      1
#> 918  -0.9657344124      1
#> 919   2.6999052212      1
#> 920   1.0749170093      1
#> 921   2.2292721037      1
#> 922   0.9977413744      1
#> 923  -1.5421633804      1
#> 924   0.7710204684      1
#> 925   1.7347748222      1
#> 926   0.3413074147      1
#> 927   2.9239324031      1
#> 928   0.3371055834      1
#> 929   1.9858403524      1
#> 930   2.6704225007      1
#> 931   1.7709425026      1
#> 932   0.2691631616      1
#> 933  -0.0939354693      1
#> 934   0.3093222879      1
#> 935   0.2577002420      1
#> 936  -0.1874762003      1
#> 937  -0.4683066887      1
#> 938   0.1833935868      1
#> 939   1.3470993780      1
#> 940  -0.0841161854      1
#> 941   0.8760264222      1
#> 942  -0.8765765400      1
#> 943   0.9877186673      1
#> 944   0.0522269909      1
#> 945   1.4329670505      1
#> 946   1.0695138959      1
#> 947   1.0026371818      1
#> 948   0.9890678007      1
#> 949   0.3344029514      1
#> 950   0.8183285470      1
#> 951   0.8208386471      1
#> 952   0.0837453280      1
#> 953   1.2847720999      1
#> 954   0.4017702526      1
#> 955  -0.7604868720      1
#> 956   0.2068190965      1
#> 957   0.9221384480      1
#> 958  -0.6340853384      1
#> 959   0.2873414465      1
#> 960   1.0164543063      1
#> 961   1.6175352188      1
#> 962   2.0168053425      1
#> 963   0.7108787413      1
#> 964  -0.5085912195      1
#> 965   0.1954748529      1
#> 966   1.9408716839      1
#> 967   1.7832462907      1
#> 968   2.6685678157      1
#> 969  -0.2484785874      1
#> 970   1.2292338731      1
#> 971   0.2094820206      1
#> 972   1.1093357040      1
#> 973   0.4831422899      1
#> 974   0.6290376266      1
#> 975   0.6506977542      1
#> 976   0.7100218014      1
#> 977   0.7087779258      1
#> 978   0.6836921179      1
#> 979   1.3037437187      1
#> 980   0.5246130813      1
#> 981   0.5489074551      1
#> 982   0.0316561539      1
#> 983   2.4933762856      1
#> 984   0.6541294128      1
#> 985   1.1729568346      1
#> 986   2.0320800627      1
#> 987   0.5916792635      1
#> 988  -0.1470140334      1
#> 989  -0.0315153849      1
#> 990   0.0396724183      1
#> 991   1.6494403134      1
#> 992  -1.3931372588      1
#> 993  -0.4946831716      1
#> 994   3.5664145481      1
#> 995  -0.7371347552      1
#> 996   0.4648564842      1
#> 997  -0.0606337209      1
#> 998   1.3159903152      1
#> 999   0.9498739511      1
#> 1000  0.0227732611      1