Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.673567671      0
#> 2     0.228434906      0
#> 3     1.916084799      0
#> 4    -0.343250548      0
#> 5     0.252596816      0
#> 6    -2.545107258      0
#> 7     1.321816637      0
#> 8    -0.626944379      0
#> 9     0.797949760      0
#> 10   -1.318396233      0
#> 11   -0.810366935      0
#> 12    0.581598374      0
#> 13    1.136020938      0
#> 14   -0.139782243      0
#> 15    1.427609637      0
#> 16    0.281147492      0
#> 17    0.731401100      0
#> 18    2.708238998      0
#> 19    0.438184064      0
#> 20    0.334106677      0
#> 21    0.004227529      0
#> 22   -0.080251587      0
#> 23   -0.184344227      0
#> 24    0.041010939      0
#> 25    0.922445453      0
#> 26   -0.032706389      0
#> 27   -1.355749162      0
#> 28    0.681871446      0
#> 29   -0.385352233      0
#> 30    1.356448585      0
#> 31    0.322025251      0
#> 32    0.252518908      0
#> 33   -0.625745099      0
#> 34   -0.347097486      0
#> 35    0.388696097      0
#> 36   -2.937380024      0
#> 37   -0.690875572      0
#> 38    0.321917099      0
#> 39    1.158888618      0
#> 40   -0.715470788      0
#> 41    1.549342469      0
#> 42    0.009948322      0
#> 43   -0.566899260      0
#> 44   -0.900955888      0
#> 45   -0.516638550      0
#> 46    0.417842280      0
#> 47    1.267028780      0
#> 48    0.363433432      0
#> 49    1.516974616      0
#> 50   -0.647413187      0
#> 51   -0.842340434      0
#> 52   -1.243103737      0
#> 53   -0.598885192      0
#> 54   -0.248210778      0
#> 55   -0.303796946      0
#> 56    0.832327107      0
#> 57    1.938026217      0
#> 58   -0.041462139      0
#> 59   -1.618976966      0
#> 60   -0.544692660      0
#> 61    0.377562164      0
#> 62    0.197833583      0
#> 63   -1.515869244      0
#> 64   -0.352270782      0
#> 65   -0.340368240      0
#> 66    0.192602688      0
#> 67   -0.211175522      0
#> 68    0.515987047      0
#> 69   -0.373283167      0
#> 70    0.914574726      0
#> 71   -1.463129374      0
#> 72    0.341201283      0
#> 73   -0.968255243      0
#> 74    0.346513584      0
#> 75    0.724843888      0
#> 76    0.377093833      0
#> 77    1.679381522      0
#> 78   -0.447935870      0
#> 79   -0.182513911      0
#> 80    0.707325563      0
#> 81   -0.696606941      0
#> 82    1.335702695      0
#> 83   -0.274604351      0
#> 84   -0.505640122      0
#> 85    1.275680661      0
#> 86   -2.155845708      0
#> 87    0.044592137      0
#> 88    0.562524201      0
#> 89    1.760418891      0
#> 90   -1.108054621      0
#> 91    0.031646666      0
#> 92    1.609510826      0
#> 93    0.584322313      0
#> 94    0.614276555      0
#> 95    0.287683444      0
#> 96   -0.707621571      0
#> 97    1.821156195      0
#> 98    0.258860620      0
#> 99    0.444464461      0
#> 100  -0.726175580      0
#> 101   0.064862004      0
#> 102  -0.674057265      0
#> 103  -1.109420938      0
#> 104   0.403817267      0
#> 105  -0.502527209      0
#> 106  -2.714086325      0
#> 107  -1.607165480      0
#> 108  -0.749760276      0
#> 109   1.590351883      0
#> 110  -1.761893474      0
#> 111   0.829568324      0
#> 112  -1.880725069      0
#> 113   0.387878724      0
#> 114  -0.889322238      0
#> 115  -0.347648772      0
#> 116   0.107057464      0
#> 117  -0.575075106      0
#> 118   0.628652949      0
#> 119  -1.843640122      0
#> 120  -1.427215740      0
#> 121  -0.114330680      0
#> 122   0.611558515      0
#> 123  -1.007419613      0
#> 124  -0.647985182      0
#> 125   0.544468041      0
#> 126  -0.785280227      0
#> 127  -0.774135298      0
#> 128   0.916539801      0
#> 129  -0.722248346      0
#> 130   0.559674707      0
#> 131   0.834494452      0
#> 132  -0.086298796      0
#> 133  -0.223160960      0
#> 134   0.975324633      0
#> 135  -2.849228398      0
#> 136   0.410683955      0
#> 137  -0.915502514      0
#> 138   0.249615251      0
#> 139  -0.566204927      0
#> 140  -1.063151863      0
#> 141   0.505617998      0
#> 142  -0.128010723      0
#> 143   0.732393202      0
#> 144  -0.151953963      0
#> 145  -0.129032903      0
#> 146  -0.444198593      0
#> 147   0.369442616      0
#> 148  -0.860719412      0
#> 149  -1.409924401      0
#> 150   1.999944195      0
#> 151   0.050612466      0
#> 152  -0.532784714      0
#> 153  -1.679446545      0
#> 154   0.651534310      0
#> 155  -0.234792371      0
#> 156  -0.984657809      0
#> 157  -0.363672273      0
#> 158   0.205558245      0
#> 159   0.683322211      0
#> 160  -1.642731623      0
#> 161   0.024500111      0
#> 162  -1.810631295      0
#> 163  -0.968495336      0
#> 164   2.025695342      0
#> 165   0.518740249      0
#> 166   1.224710938      0
#> 167  -0.628486596      0
#> 168   0.058523066      0
#> 169   0.486273119      0
#> 170  -0.188620739      0
#> 171   1.616369645      0
#> 172  -3.287194118      0
#> 173   0.174411915      0
#> 174  -0.171834354      0
#> 175   1.648192576      0
#> 176   0.093712935      0
#> 177  -1.298207849      0
#> 178  -0.893620336      0
#> 179  -2.105152218      0
#> 180  -0.245171695      0
#> 181   1.090883563      0
#> 182  -0.372588878      0
#> 183   1.430896150      0
#> 184  -0.346284215      0
#> 185   0.571492162      0
#> 186   0.517135068      0
#> 187   0.066400651      0
#> 188   0.075841295      0
#> 189  -0.004871938      0
#> 190   1.515388387      0
#> 191   0.794392405      0
#> 192  -0.874471073      0
#> 193  -0.748470774      0
#> 194  -2.002700521      0
#> 195   0.613231051      0
#> 196  -0.968629742      0
#> 197   0.594152874      0
#> 198  -0.684436628      0
#> 199  -0.007397337      0
#> 200   1.053298494      0
#> 201  -0.785029632      0
#> 202  -1.128538686      0
#> 203  -0.390505377      0
#> 204  -0.098104293      0
#> 205   1.538607367      0
#> 206  -0.082837186      0
#> 207  -0.295619719      0
#> 208   0.175269568      0
#> 209   1.198695668      0
#> 210   0.374628730      0
#> 211  -1.318408687      0
#> 212  -0.344476470      0
#> 213  -1.575189810      0
#> 214  -1.919472425      0
#> 215  -0.286760351      0
#> 216  -0.923965828      0
#> 217  -1.102117519      0
#> 218   0.101034763      0
#> 219   0.856352045      0
#> 220   0.300071425      0
#> 221   2.062402313      0
#> 222  -0.627385154      0
#> 223   2.330324715      0
#> 224   1.235053422      0
#> 225   0.183162156      0
#> 226   1.499424988      0
#> 227   0.412647754      0
#> 228  -1.322652474      0
#> 229  -0.257148292      0
#> 230   0.964663050      0
#> 231   0.596839534      0
#> 232  -0.926332370      0
#> 233   0.077470740      0
#> 234   0.805647871      0
#> 235   0.528192964      0
#> 236  -1.959189585      0
#> 237   0.782530784      0
#> 238   0.246143756      0
#> 239  -0.332681459      0
#> 240   0.881934213      0
#> 241  -2.011278934      0
#> 242  -0.876601281      0
#> 243   1.322391837      0
#> 244   0.641825382      0
#> 245  -1.751240580      0
#> 246  -0.511172162      0
#> 247  -1.619298742      0
#> 248   1.480993096      0
#> 249   0.456896283      0
#> 250   0.887225554      0
#> 251  -1.343249557      0
#> 252   0.056236091      0
#> 253  -1.284954677      0
#> 254   0.287941025      0
#> 255  -0.864115651      0
#> 256   1.055317205      0
#> 257   0.140162757      0
#> 258  -0.100092382      0
#> 259   1.130246261      0
#> 260   0.338542763      0
#> 261  -0.878203914      0
#> 262  -0.660751234      0
#> 263  -0.886763578      0
#> 264   0.591549910      0
#> 265   0.134340010      0
#> 266  -1.662746167      0
#> 267  -0.350169938      0
#> 268   0.932980588      0
#> 269   1.719157950      0
#> 270  -0.015885254      0
#> 271   0.798275254      0
#> 272  -0.639987773      0
#> 273   1.048625569      0
#> 274  -1.188413151      0
#> 275   1.553521881      0
#> 276  -2.053743406      0
#> 277  -0.424051822      0
#> 278   0.748530313      0
#> 279  -0.944463677      0
#> 280   0.622786455      0
#> 281   1.375955233      0
#> 282   0.829994983      0
#> 283  -0.662456137      0
#> 284   0.146038342      0
#> 285  -1.088318800      0
#> 286  -2.534007476      0
#> 287   1.508765346      0
#> 288   0.485719215      0
#> 289   0.538550313      0
#> 290   1.006701273      0
#> 291  -0.310422603      0
#> 292   0.737427242      0
#> 293   1.234793158      0
#> 294   0.601119491      0
#> 295  -0.863205027      0
#> 296   0.088903027      0
#> 297  -0.346234967      0
#> 298   0.969175673      0
#> 299  -0.957972542      0
#> 300   0.445378827      0
#> 301   2.100795481      0
#> 302   1.300503624      0
#> 303  -0.959422308      0
#> 304  -1.877675792      0
#> 305   0.859081541      0
#> 306  -0.358829266      0
#> 307   0.900628370      0
#> 308   1.898740160      0
#> 309   0.979721223      0
#> 310   0.042667031      0
#> 311   0.167139586      0
#> 312  -0.890343771      0
#> 313  -0.751771868      0
#> 314  -1.169680809      0
#> 315  -1.998134250      0
#> 316  -0.368277197      0
#> 317   1.365095616      0
#> 318  -1.293296509      0
#> 319  -0.119406641      0
#> 320  -0.661752921      0
#> 321   0.496636834      0
#> 322   0.134542418      0
#> 323   0.478517702      0
#> 324  -0.373367967      0
#> 325  -0.007665031      0
#> 326  -1.412067711      0
#> 327  -1.242107197      0
#> 328  -0.967539900      0
#> 329   0.921889541      0
#> 330  -0.812909347      0
#> 331   0.144273426      0
#> 332  -0.735197771      0
#> 333  -0.544138353      0
#> 334   1.297012237      0
#> 335  -0.566387053      0
#> 336  -0.098333561      0
#> 337   0.242413452      0
#> 338  -0.830026541      0
#> 339   0.981622812      0
#> 340  -0.343971594      0
#> 341  -0.306279137      0
#> 342  -0.374526426      0
#> 343  -1.178414330      0
#> 344   0.976707501      0
#> 345   0.354288302      0
#> 346  -1.485261992      0
#> 347  -0.321786784      0
#> 348  -1.711346721      0
#> 349   0.964876310      0
#> 350   0.453324618      0
#> 351  -1.387054670      0
#> 352   0.499862570      0
#> 353   1.962592433      0
#> 354   1.403262487      0
#> 355  -0.196445175      0
#> 356  -0.991997111      0
#> 357  -0.360255391      0
#> 358   1.216142745      0
#> 359  -1.300484195      0
#> 360  -0.482989051      0
#> 361   1.144056700      0
#> 362  -0.483315631      0
#> 363  -0.221633908      0
#> 364   0.463313258      0
#> 365   0.155533224      0
#> 366  -0.280023385      0
#> 367   0.993772375      0
#> 368   0.266085314      0
#> 369  -0.778662232      0
#> 370  -1.414612475      0
#> 371  -1.004019103      0
#> 372  -0.414127971      0
#> 373   2.610349227      0
#> 374   0.550307340      0
#> 375  -0.487911158      0
#> 376  -2.172733230      0
#> 377   0.697710939      0
#> 378   1.755592192      0
#> 379  -1.276340693      0
#> 380   2.303302036      0
#> 381  -2.646632049      0
#> 382   1.297811980      0
#> 383  -0.111866522      0
#> 384  -1.479479130      0
#> 385   0.449807616      0
#> 386   0.361004370      0
#> 387  -0.589805596      0
#> 388  -1.288010632      0
#> 389   0.843836658      0
#> 390   1.261272540      0
#> 391   0.391546560      0
#> 392   1.488798335      0
#> 393  -0.418964515      0
#> 394  -0.267612300      0
#> 395  -0.067586149      0
#> 396  -0.001079797      0
#> 397   0.251566289      0
#> 398  -0.415290020      0
#> 399   0.501632372      0
#> 400   0.530540246      0
#> 401  -1.093762656      0
#> 402  -0.684534515      0
#> 403  -0.946720121      0
#> 404   0.374246859      0
#> 405  -0.214874056      0
#> 406   0.747732981      0
#> 407   1.534791817      0
#> 408   1.251551751      0
#> 409   0.714959593      0
#> 410  -0.391008921      0
#> 411   0.679823773      0
#> 412  -0.024220825      0
#> 413  -1.167258474      0
#> 414   0.879667529      0
#> 415  -0.455908833      0
#> 416  -0.037726604      0
#> 417   0.580125854      0
#> 418   0.761029420      0
#> 419   0.319602399      0
#> 420  -1.501829562      0
#> 421   1.892600475      0
#> 422  -0.579404162      0
#> 423  -0.352590871      0
#> 424   0.754447564      0
#> 425   0.384906670      0
#> 426  -1.842309527      0
#> 427  -0.796833977      0
#> 428   0.017737386      0
#> 429   1.119189729      0
#> 430   0.969576216      0
#> 431  -1.413873837      0
#> 432   0.408409361      0
#> 433  -0.568959815      0
#> 434   0.101055528      0
#> 435  -0.395193599      0
#> 436  -0.548820265      0
#> 437   0.982798688      0
#> 438  -1.082375169      0
#> 439  -1.127434883      0
#> 440  -0.588367837      0
#> 441  -0.076298518      0
#> 442  -0.418882363      0
#> 443  -0.999044058      0
#> 444  -0.411750942      0
#> 445   1.084528288      0
#> 446  -0.257964057      0
#> 447   0.008277720      0
#> 448  -0.906199793      0
#> 449   0.010205759      0
#> 450  -1.261837014      0
#> 451  -1.308569087      0
#> 452  -2.100850301      0
#> 453  -0.095552347      0
#> 454  -1.314361489      0
#> 455  -0.485634798      0
#> 456   0.239270279      0
#> 457   0.757602492      0
#> 458   0.102963274      0
#> 459   0.784915552      0
#> 460   1.079731538      0
#> 461  -1.527227727      0
#> 462   2.714682886      0
#> 463  -1.563592575      0
#> 464  -0.719311558      0
#> 465  -1.789706972      0
#> 466   0.399478933      0
#> 467   0.056692765      0
#> 468   1.421819263      0
#> 469  -0.411461194      0
#> 470  -2.318765417      0
#> 471   1.437172902      0
#> 472  -0.285914335      0
#> 473  -0.861670514      0
#> 474   0.763893193      0
#> 475   0.314758619      0
#> 476   0.204765871      0
#> 477   1.440882687      0
#> 478   0.438294320      0
#> 479   0.394232792      0
#> 480  -1.107803924      0
#> 481  -1.183544472      0
#> 482  -0.207434682      0
#> 483  -0.105941734      0
#> 484   1.581133078      0
#> 485   0.993663785      0
#> 486   0.899913195      0
#> 487  -0.824413594      0
#> 488  -1.002402280      0
#> 489  -0.206810232      0
#> 490  -1.439711310      0
#> 491   1.243639707      0
#> 492  -1.186257484      0
#> 493  -1.260172417      0
#> 494  -0.188059804      0
#> 495   0.141099985      0
#> 496  -1.022123368      0
#> 497   0.654507090      0
#> 498  -0.388860167      0
#> 499  -1.877618042      0
#> 500  -0.248849921      0
#> 501  -1.128054214      0
#> 502  -0.092339036      0
#> 503  -0.838316476      0
#> 504   0.814979857      0
#> 505   0.389467599      0
#> 506  -0.010859145      0
#> 507  -0.582795061      0
#> 508   1.021576374      0
#> 509  -1.896192687      0
#> 510   0.137846117      0
#> 511  -1.850667834      0
#> 512  -0.831331496      0
#> 513   1.470232696      0
#> 514  -0.438572767      0
#> 515   1.273036329      0
#> 516   1.351868651      0
#> 517   0.355776230      0
#> 518  -1.554514070      0
#> 519  -0.038984675      0
#> 520  -1.376874162      0
#> 521   0.709503175      0
#> 522  -1.200946121      0
#> 523  -0.506480128      0
#> 524  -1.081055283      0
#> 525  -0.221293701      0
#> 526   0.079651796      0
#> 527   0.184911044      0
#> 528  -0.374988586      0
#> 529   1.336663185      0
#> 530  -1.192912194      0
#> 531  -0.705402686      0
#> 532   0.935570924      0
#> 533  -0.972847570      0
#> 534   1.901243901      0
#> 535   1.186130917      0
#> 536  -0.552710630      0
#> 537  -1.451259428      0
#> 538   0.485091241      0
#> 539  -0.419321077      0
#> 540  -2.224275242      0
#> 541   0.152854595      0
#> 542  -0.259460093      0
#> 543  -1.790081962      0
#> 544  -0.357836217      0
#> 545   0.120375854      0
#> 546  -1.451890783      0
#> 547   1.473174206      0
#> 548  -0.440069038      0
#> 549   0.338560473      0
#> 550  -0.646519396      0
#> 551   0.356415218      0
#> 552   1.721171990      0
#> 553  -0.928930050      0
#> 554   0.937889124      0
#> 555   0.304672075      0
#> 556  -0.394177697      0
#> 557  -0.542364914      0
#> 558   1.492905013      0
#> 559   1.282668647      0
#> 560   0.012065242      0
#> 561  -1.524386523      0
#> 562  -1.218793955      0
#> 563   0.062115603      0
#> 564  -0.851193046      0
#> 565   0.833983868      0
#> 566   1.015604086      0
#> 567  -0.279303770      0
#> 568   0.103244103      0
#> 569   0.574938316      0
#> 570   0.250484957      0
#> 571   0.432603230      0
#> 572   1.272256476      0
#> 573   0.250685630      0
#> 574   0.425829026      0
#> 575   1.088025394      0
#> 576   0.780444323      0
#> 577  -0.507451996      0
#> 578   0.540933977      0
#> 579   0.436040552      0
#> 580  -1.630905171      0
#> 581  -0.632856680      0
#> 582   1.421361812      0
#> 583   0.796252227      0
#> 584   0.703457163      0
#> 585   0.185626775      0
#> 586  -2.393672137      0
#> 587  -1.601891582      0
#> 588  -1.350341298      0
#> 589  -0.407295742      0
#> 590  -1.239378067      0
#> 591   0.011082266      0
#> 592   1.170214128      0
#> 593  -0.174897570      0
#> 594  -0.727725021      0
#> 595   0.176988472      0
#> 596   0.289416082      0
#> 597   0.210727975      0
#> 598   0.082297456      0
#> 599   0.444468377      0
#> 600   0.061804587      0
#> 601   1.062929448      0
#> 602  -0.090390075      0
#> 603   1.038425388      0
#> 604  -0.264385061      0
#> 605  -0.950929251      0
#> 606   0.142434771      0
#> 607   1.379157835      0
#> 608   0.610387760      0
#> 609  -0.651518236      0
#> 610   1.398752536      0
#> 611  -0.167328617      0
#> 612  -1.482810571      0
#> 613   0.445486079      0
#> 614   0.563549057      0
#> 615   0.670464838      0
#> 616   0.415626196      0
#> 617  -0.874815045      0
#> 618  -0.645353120      0
#> 619  -2.493713538      0
#> 620   0.099323099      0
#> 621   0.130566629      0
#> 622   0.073454415      0
#> 623   0.355236145      0
#> 624   1.883267944      0
#> 625  -0.908188689      0
#> 626  -1.418043984      0
#> 627  -1.531065732      0
#> 628  -0.025160413      0
#> 629  -0.099907946      0
#> 630   2.388301868      0
#> 631  -0.554540958      0
#> 632   0.807844398      0
#> 633   0.916604985      0
#> 634   1.846867000      0
#> 635  -1.819745514      0
#> 636   1.699654526      0
#> 637   0.743800532      0
#> 638  -0.392611152      0
#> 639  -0.635378897      0
#> 640   2.327222966      0
#> 641   0.561699680      0
#> 642  -0.232061383      0
#> 643  -0.987922862      0
#> 644  -0.159028557      0
#> 645   0.502935090      0
#> 646   0.866595085      0
#> 647   0.043169562      0
#> 648  -0.277854243      0
#> 649  -0.198526876      0
#> 650  -1.390140976      0
#> 651   0.136874325      0
#> 652  -0.186557300      0
#> 653  -0.465128346      0
#> 654  -1.023375125      0
#> 655  -0.436723337      0
#> 656  -0.057834234      0
#> 657  -0.800603743      0
#> 658  -1.709777047      0
#> 659  -0.610865861      0
#> 660  -0.534870033      0
#> 661   0.887889881      0
#> 662  -0.577739953      0
#> 663  -0.833148330      0
#> 664   0.117873271      0
#> 665  -2.380315699      0
#> 666   0.507878703      0
#> 667   1.562827851      0
#> 668   1.310139060      0
#> 669   0.097536098      0
#> 670  -0.086926678      0
#> 671  -0.865282398      0
#> 672   1.786111991      0
#> 673  -0.135379189      0
#> 674   2.080203454      0
#> 675   0.520964597      0
#> 676   0.215872692      0
#> 677  -0.460539611      0
#> 678  -0.476160606      0
#> 679  -0.545534752      0
#> 680  -1.134891233      0
#> 681   0.505930957      0
#> 682  -0.761722469      0
#> 683   1.525059052      0
#> 684  -0.779474492      0
#> 685  -0.721848678      0
#> 686  -1.134083047      0
#> 687   1.121034002      0
#> 688  -0.523010180      0
#> 689  -0.338903122      0
#> 690   0.769261039      0
#> 691   0.301341028      0
#> 692  -1.252421623      0
#> 693   0.700261561      0
#> 694   0.005326256      0
#> 695  -0.499075792      0
#> 696  -1.784030133      0
#> 697   0.863634133      0
#> 698   0.280965636      0
#> 699   0.591558653      0
#> 700   0.588500201      0
#> 701  -0.950930567      0
#> 702  -2.776948888      0
#> 703  -1.256690069      0
#> 704  -2.075564377      0
#> 705  -1.269169191      0
#> 706  -0.729692799      0
#> 707  -0.306655719      0
#> 708   2.490191453      0
#> 709   0.809923626      0
#> 710   1.886218940      0
#> 711  -0.689268962      0
#> 712  -1.534153392      0
#> 713  -1.538671818      0
#> 714  -0.310153310      0
#> 715  -0.916363647      0
#> 716   0.926682394      0
#> 717   1.614423914      0
#> 718   1.254649181      0
#> 719   2.189062783      0
#> 720  -1.012964344      0
#> 721   0.261499955      0
#> 722   1.595763891      0
#> 723   0.379368158      0
#> 724  -0.376108992      0
#> 725   1.360661622      0
#> 726  -1.072552860      0
#> 727  -0.916766128      0
#> 728  -1.045819185      0
#> 729   0.274180733      0
#> 730  -2.660012726      0
#> 731  -0.998886416      0
#> 732   0.471073674      0
#> 733   0.491213273      0
#> 734  -0.797798243      0
#> 735  -0.287335912      0
#> 736   0.724477299      0
#> 737   0.305325697      0
#> 738  -0.886640114      0
#> 739  -1.503353767      0
#> 740  -0.865618137      0
#> 741   0.024527336      0
#> 742   0.767010539      0
#> 743   0.452463360      0
#> 744  -0.930541558      0
#> 745   1.594706507      0
#> 746   1.140173937      0
#> 747   0.345387975      0
#> 748  -1.210055291      0
#> 749  -0.791089658      0
#> 750  -0.903775831      0
#> 751  -1.265494376      0
#> 752  -1.783606036      0
#> 753   0.338402629      0
#> 754   0.531102259      0
#> 755   0.406400028      0
#> 756  -0.522666311      0
#> 757   0.441709847      0
#> 758  -1.336349461      0
#> 759  -1.477569043      0
#> 760   0.246444071      0
#> 761  -0.196526330      0
#> 762   0.059911750      0
#> 763  -0.635049317      0
#> 764   0.426799291      0
#> 765   1.066522044      0
#> 766  -0.650025242      0
#> 767   0.044384853      0
#> 768   0.690572936      0
#> 769   0.649930545      0
#> 770   1.475831468      0
#> 771   1.938561300      0
#> 772   0.405065867      0
#> 773   0.450815836      0
#> 774   0.861949047      0
#> 775  -0.293579787      0
#> 776  -0.634743698      0
#> 777  -0.546955134      0
#> 778   0.395019402      0
#> 779  -1.106038518      0
#> 780  -0.125507198      0
#> 781   0.405652532      0
#> 782  -1.401547252      0
#> 783  -1.030980399      0
#> 784  -0.177695892      0
#> 785   1.111151100      0
#> 786  -0.204679663      0
#> 787  -0.752018895      0
#> 788  -0.580713534      0
#> 789   0.510078172      0
#> 790  -0.810710175      0
#> 791  -0.744462911      0
#> 792  -1.046840266      0
#> 793   0.786688192      0
#> 794  -0.194909458      0
#> 795   0.774422066      0
#> 796   0.097854105      0
#> 797  -0.688374014      0
#> 798   0.115754281      0
#> 799   0.082249166      0
#> 800  -0.886374975      0
#> 801   0.453548561      0
#> 802   0.469227440      0
#> 803   0.640555302      0
#> 804   0.800788047      0
#> 805   1.928564916      0
#> 806   1.276590636      0
#> 807  -0.042564764      0
#> 808  -0.510672593      0
#> 809   0.226281528      0
#> 810  -0.547680911      0
#> 811   0.714440669      0
#> 812  -0.188488485      0
#> 813  -0.514706590      0
#> 814   0.618439186      0
#> 815  -0.275406291      0
#> 816   1.389756004      0
#> 817  -0.533151951      0
#> 818  -0.668852892      0
#> 819   1.722231330      0
#> 820   1.228793067      0
#> 821  -0.701717273      0
#> 822  -0.505467372      0
#> 823   0.695928692      0
#> 824  -1.661694013      0
#> 825   0.261112998      0
#> 826  -0.522406453      0
#> 827   1.644568882      0
#> 828  -0.696378929      0
#> 829  -0.457543780      0
#> 830  -0.026809688      0
#> 831  -0.941356350      0
#> 832  -1.317960719      0
#> 833   0.140552067      0
#> 834   0.567319777      0
#> 835   0.133895933      0
#> 836   0.210940976      0
#> 837  -1.684957561      0
#> 838  -0.243510772      0
#> 839   0.997564158      0
#> 840  -0.151748032      0
#> 841  -0.262069489      0
#> 842   1.123093452      0
#> 843   0.210684397      0
#> 844  -0.159841809      0
#> 845  -0.204603599      0
#> 846  -0.945060959      0
#> 847   1.312285833      0
#> 848   0.977827139      0
#> 849  -2.173159800      0
#> 850  -2.250631156      0
#> 851   0.270926243      0
#> 852   1.348940199      0
#> 853  -1.250330119      0
#> 854  -1.312858894      0
#> 855   0.058129913      0
#> 856  -0.443852371      0
#> 857   0.448927202      0
#> 858  -1.444259297      0
#> 859  -0.309076850      0
#> 860  -1.050679168      0
#> 861   0.072951452      0
#> 862   0.893673649      0
#> 863   0.312462658      0
#> 864  -1.292728785      0
#> 865  -0.129874912      0
#> 866  -0.666669541      0
#> 867   0.140386267      0
#> 868  -1.512445757      0
#> 869   0.518943336      0
#> 870   0.060465168      0
#> 871   0.667127815      0
#> 872  -0.582592686      0
#> 873  -0.409818634      0
#> 874   0.384218274      0
#> 875   0.120357209      0
#> 876  -1.316045698      0
#> 877  -0.946788742      0
#> 878  -0.859697262      0
#> 879   1.226433729      0
#> 880   0.472230619      0
#> 881  -2.496235110      0
#> 882  -0.310912024      0
#> 883  -1.159361678      0
#> 884  -0.687467914      0
#> 885   0.673207079      0
#> 886   1.058267080      0
#> 887   0.165396677      0
#> 888   0.982311215      0
#> 889   1.025657458      0
#> 890   0.398761490      0
#> 891  -1.174056145      0
#> 892   0.640394950      0
#> 893   0.876115011      0
#> 894  -0.545401526      0
#> 895   0.511471385      0
#> 896  -0.707659147      0
#> 897   0.025739576      0
#> 898   1.679277070      1
#> 899  -1.228677510      1
#> 900   2.377173778      1
#> 901   0.495058681      1
#> 902   0.473074901      1
#> 903   1.662504142      1
#> 904   2.353142073      1
#> 905  -1.063163713      1
#> 906   2.160242513      1
#> 907   2.062320087      1
#> 908  -0.086201252      1
#> 909   2.097869332      1
#> 910  -0.766902521      1
#> 911   1.522760303      1
#> 912   0.736291665      1
#> 913  -0.243814454      1
#> 914   1.140749061      1
#> 915   0.072102269      1
#> 916   2.263075498      1
#> 917   0.837695173      1
#> 918  -0.628477997      1
#> 919  -1.238272679      1
#> 920   1.503496442      1
#> 921   1.148559730      1
#> 922   1.112600219      1
#> 923   2.534947610      1
#> 924   0.302228969      1
#> 925  -0.558496986      1
#> 926  -1.456347173      1
#> 927   0.284621015      1
#> 928   2.348323718      1
#> 929   1.819509948      1
#> 930   0.915774577      1
#> 931  -0.574940283      1
#> 932   1.406831105      1
#> 933   0.020780416      1
#> 934   0.216925828      1
#> 935  -0.246179710      1
#> 936  -0.033253594      1
#> 937  -0.158249468      1
#> 938   2.435987954      1
#> 939  -0.029248864      1
#> 940   2.190036680      1
#> 941   0.920109609      1
#> 942   0.509573822      1
#> 943   1.785720929      1
#> 944  -0.452982250      1
#> 945   3.113633276      1
#> 946  -0.050660382      1
#> 947  -0.509719575      1
#> 948   1.069315141      1
#> 949   0.286404636      1
#> 950   0.004490778      1
#> 951   1.618049709      1
#> 952   1.889143802      1
#> 953  -0.754572739      1
#> 954   1.060979509      1
#> 955   0.832959971      1
#> 956   1.698746291      1
#> 957   0.304822034      1
#> 958   1.605603918      1
#> 959  -0.142453909      1
#> 960   0.429819972      1
#> 961   1.534495814      1
#> 962   0.773394728      1
#> 963   1.691268311      1
#> 964   2.626422850      1
#> 965   0.967376305      1
#> 966  -0.300784615      1
#> 967   2.602750776      1
#> 968   0.778672741      1
#> 969   0.578674009      1
#> 970   0.223541509      1
#> 971   2.434207922      1
#> 972   1.519653063      1
#> 973   1.515412593      1
#> 974   1.381243922      1
#> 975   1.244899447      1
#> 976   1.302884139      1
#> 977   0.041999467      1
#> 978   1.240827833      1
#> 979  -0.565491296      1
#> 980   0.157759674      1
#> 981  -0.035385421      1
#> 982  -0.444705103      1
#> 983  -2.197577685      1
#> 984   2.062008787      1
#> 985   0.741929141      1
#> 986  -0.721134533      1
#> 987   1.475873387      1
#> 988   0.574344762      1
#> 989   0.388226764      1
#> 990  -0.123705037      1
#> 991   1.636591535      1
#> 992   1.416741843      1
#> 993   1.284456418      1
#> 994   0.081140741      1
#> 995  -0.453748937      1
#> 996   0.065249722      1
#> 997   0.867518883      1
#> 998   2.237072238      1
#> 999   1.767562371      1
#> 1000  1.539654632      1