Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -1.2115858490      0
#> 2    -1.6267287151      0
#> 3     0.6857206888      0
#> 4     1.5547116216      0
#> 5    -0.4527313520      0
#> 6    -0.5941122337      0
#> 7    -0.9669490569      0
#> 8     1.2351120558      0
#> 9    -0.1056217990      0
#> 10   -0.6163036647      0
#> 11    0.4070574304      0
#> 12   -0.7449800185      0
#> 13   -0.0380926227      0
#> 14    1.5787489021      0
#> 15    0.1670033503      0
#> 16   -0.6380365933      0
#> 17    0.5634304224      0
#> 18   -0.7188797382      0
#> 19   -0.6488629712      0
#> 20   -0.3583316688      0
#> 21    0.7603917145      0
#> 22   -0.8817404271      0
#> 23    1.2411058133      0
#> 24   -1.6372161941      0
#> 25   -0.9442218713      0
#> 26    0.0305818927      0
#> 27    1.1777398534      0
#> 28   -0.5670192600      0
#> 29    0.5746824668      0
#> 30   -1.3357080056      0
#> 31   -1.7842193793      0
#> 32   -1.0093085075      0
#> 33    0.7650916672      0
#> 34   -1.0397450911      0
#> 35   -0.8752276280      0
#> 36    2.0020617632      0
#> 37    0.8578540781      0
#> 38   -0.1051195950      0
#> 39   -2.3223888183      0
#> 40    0.3048353037      0
#> 41   -1.1687982495      0
#> 42    1.6977258796      0
#> 43   -0.0484904083      0
#> 44   -1.4531886564      0
#> 45    1.1648708538      0
#> 46   -0.4583758768      0
#> 47   -0.3197764740      0
#> 48   -0.0692007265      0
#> 49    0.2212384574      0
#> 50   -0.2910302394      0
#> 51   -0.0637977130      0
#> 52   -0.3413734791      0
#> 53   -0.9224221216      0
#> 54   -0.2833851053      0
#> 55   -0.4410813974      0
#> 56    0.8709268949      0
#> 57    0.0656109990      0
#> 58   -0.4284694492      0
#> 59    0.0971459783      0
#> 60    0.3272730894      0
#> 61    0.6179312484      0
#> 62    0.7209786176      0
#> 63    1.7108344579      0
#> 64    0.1363460630      0
#> 65    1.0105762826      0
#> 66    1.4101198625      0
#> 67    0.7699967308      0
#> 68   -1.0856920713      0
#> 69   -0.0771405656      0
#> 70    0.3488765319      0
#> 71   -0.1295311746      0
#> 72    2.1014784482      0
#> 73   -0.7365060055      0
#> 74   -1.3194742371      0
#> 75    0.0171393046      0
#> 76    0.2233603760      0
#> 77   -0.8118126372      0
#> 78   -0.2175274353      0
#> 79    0.5951280037      0
#> 80   -0.1968680680      0
#> 81    1.8090914050      0
#> 82    1.3875531922      0
#> 83   -1.3789493635      0
#> 84    1.2226136226      0
#> 85    1.1651909432      0
#> 86   -1.1916435509      0
#> 87   -0.5771321355      0
#> 88   -0.8893435649      0
#> 89   -0.0265119432      0
#> 90   -0.5608303615      0
#> 91    0.3250657721      0
#> 92    0.3500998612      0
#> 93    1.2808668146      0
#> 94    0.4921851329      0
#> 95    0.1311872682      0
#> 96    0.5078984817      0
#> 97   -0.2635190917      0
#> 98   -0.9444856652      0
#> 99    1.0813812932      0
#> 100  -0.4083162667      0
#> 101  -1.4656904267      0
#> 102   0.1018961527      0
#> 103  -0.0366050720      0
#> 104  -0.3865984804      0
#> 105   1.0012552240      0
#> 106   0.1352790291      0
#> 107   0.1594065966      0
#> 108   2.5604275806      0
#> 109  -1.2909468863      0
#> 110   1.3564902710      0
#> 111  -0.1983771153      0
#> 112  -1.2891203502      0
#> 113  -0.0749420706      0
#> 114  -0.0012319917      0
#> 115  -0.4590792771      0
#> 116  -0.0145450273      0
#> 117   0.1460710586      0
#> 118  -1.0816260958      0
#> 119  -0.5182429707      0
#> 120   1.3877403085      0
#> 121  -0.1943508461      0
#> 122  -0.0932497719      0
#> 123  -0.3877323861      0
#> 124  -1.3951115032      0
#> 125   1.1111043274      0
#> 126   0.7835910385      0
#> 127   1.4961749254      0
#> 128  -1.3925903713      0
#> 129  -0.0606936163      0
#> 130  -0.5861757672      0
#> 131   0.6389268142      0
#> 132  -0.2508656827      0
#> 133  -1.6027900724      0
#> 134  -0.3798541461      0
#> 135  -0.4746024066      0
#> 136  -2.5583582801      0
#> 137  -1.0673153626      0
#> 138  -0.9836965501      0
#> 139  -0.0352265895      0
#> 140  -1.2840032188      0
#> 141   1.0886542625      0
#> 142   0.9455016604      0
#> 143   0.3282123634      0
#> 144  -1.6315252129      0
#> 145   2.0068546105      0
#> 146   1.6478673222      0
#> 147  -0.9233094512      0
#> 148  -0.1665302988      0
#> 149  -1.3346547430      0
#> 150  -1.8032376222      0
#> 151  -0.8636366785      0
#> 152  -0.3228885998      0
#> 153   0.0203895451      0
#> 154   0.2023589412      0
#> 155   1.0462750286      0
#> 156  -0.5851908767      0
#> 157  -0.0033993510      0
#> 158  -0.9708393319      0
#> 159  -0.7731881644      0
#> 160   0.5660353384      0
#> 161  -0.1623175078      0
#> 162  -1.4227234552      0
#> 163  -0.9226216710      0
#> 164   0.5676502336      0
#> 165  -0.8849997433      0
#> 166   0.2008724010      0
#> 167   0.0232150792      0
#> 168   1.2737303184      0
#> 169   0.0581135772      0
#> 170   0.8873566936      0
#> 171  -1.0358630653      0
#> 172   0.4786311825      0
#> 173   0.7846341574      0
#> 174   2.5218076338      0
#> 175  -0.3430295169      0
#> 176   0.3327603413      0
#> 177  -0.9009255504      0
#> 178   0.2623992738      0
#> 179   0.1868567875      0
#> 180   0.1014030456      0
#> 181  -1.1085680954      0
#> 182   0.9639060916      0
#> 183  -0.7826585137      0
#> 184  -0.9292017766      0
#> 185  -0.0651993106      0
#> 186  -0.2223483538      0
#> 187   1.6539750543      0
#> 188   0.8328170937      0
#> 189  -0.7177205384      0
#> 190  -0.5780694992      0
#> 191   1.2966187148      0
#> 192  -0.5790899084      0
#> 193   0.2801647297      0
#> 194  -0.4543765959      0
#> 195   0.8993018279      0
#> 196   1.0566798765      0
#> 197   0.4200942561      0
#> 198  -2.8962318236      0
#> 199   0.4125518245      0
#> 200   0.2189590020      0
#> 201   0.0981117617      0
#> 202   0.5774082226      0
#> 203   1.0916238066      0
#> 204   1.2588686078      0
#> 205  -0.2962376912      0
#> 206   0.5083233645      0
#> 207  -0.4459937900      0
#> 208   0.6053143876      0
#> 209  -1.2727083824      0
#> 210   1.0904896063      0
#> 211   2.6746275430      0
#> 212  -0.7074123530      0
#> 213  -0.4948193889      0
#> 214  -0.3402603529      0
#> 215   1.5138164141      0
#> 216   1.4108275469      0
#> 217   2.1511096229      0
#> 218  -0.7796572723      0
#> 219  -0.7356200876      0
#> 220   1.3986626055      0
#> 221   0.6273372088      0
#> 222  -0.6290907619      0
#> 223  -0.6069849240      0
#> 224  -0.7416675513      0
#> 225   0.2856599867      0
#> 226   0.3273993642      0
#> 227   0.4030570994      0
#> 228  -0.2097922689      0
#> 229   1.1054106466      0
#> 230   0.5616796675      0
#> 231  -1.7708993523      0
#> 232  -0.1825195292      0
#> 233   0.0640191815      0
#> 234  -0.2607896741      0
#> 235  -0.4524248278      0
#> 236   0.9617970085      0
#> 237   0.2481046821      0
#> 238  -1.9299039268      0
#> 239   0.1169796403      0
#> 240  -0.7566327306      0
#> 241   1.9482139769      0
#> 242   0.0743792438      0
#> 243   1.3722163769      0
#> 244   0.4226271379      0
#> 245  -1.3967800343      0
#> 246   0.6640384529      0
#> 247  -1.5203194794      0
#> 248  -0.5122936183      0
#> 249   2.6354864738      0
#> 250   1.7509053274      0
#> 251  -1.6294083806      0
#> 252  -0.0253339111      0
#> 253  -0.4422406476      0
#> 254   0.4851438332      0
#> 255  -1.3061175215      0
#> 256  -0.8643008285      0
#> 257  -1.0001995814      0
#> 258   0.4988731464      0
#> 259   0.2380595335      0
#> 260   1.9467335384      0
#> 261  -1.7524919936      0
#> 262  -1.0394991691      0
#> 263  -0.0539628371      0
#> 264  -0.9130389430      0
#> 265   0.0647855601      0
#> 266   1.4559578415      0
#> 267  -0.8708894871      0
#> 268   0.0866610590      0
#> 269   0.8242935573      0
#> 270  -1.1178237089      0
#> 271   0.8332835738      0
#> 272   0.2994037357      0
#> 273  -0.2826001550      0
#> 274   0.2688807273      0
#> 275  -0.8440951146      0
#> 276   0.7200941350      0
#> 277   0.7761409772      0
#> 278  -0.1068004090      0
#> 279   1.8226348931      0
#> 280  -0.7256862561      0
#> 281   0.8012845569      0
#> 282   0.1950261919      0
#> 283  -0.3527893931      0
#> 284  -0.7153111290      0
#> 285  -0.4345911070      0
#> 286   1.4927267871      0
#> 287  -0.3468252942      0
#> 288  -0.2460463241      0
#> 289   2.5036498920      0
#> 290   0.6795773693      0
#> 291  -0.7346089096      0
#> 292  -1.0403865210      0
#> 293   0.1967444535      0
#> 294  -0.1209247170      0
#> 295   0.2202695334      0
#> 296  -0.8905120348      0
#> 297  -0.4257360825      0
#> 298  -1.5680644300      0
#> 299   1.2494715044      0
#> 300  -1.6204070578      0
#> 301   0.8198318042      0
#> 302  -0.1896862817      0
#> 303  -0.5205229422      0
#> 304  -0.6005011030      0
#> 305   2.7835156448      0
#> 306  -1.4637012286      0
#> 307  -0.0001204667      0
#> 308   0.2120943068      0
#> 309  -0.1270419278      0
#> 310  -0.0901762096      0
#> 311   0.7637193094      0
#> 312  -1.1523734603      0
#> 313  -2.0307643268      0
#> 314   0.2273351001      0
#> 315   0.5862085796      0
#> 316  -0.1724216040      0
#> 317   0.3474638050      0
#> 318   1.0306624627      0
#> 319  -0.6682823805      0
#> 320   0.4617132480      0
#> 321  -2.6330632189      0
#> 322   0.1561330081      0
#> 323  -0.0421954325      0
#> 324   0.0218176290      0
#> 325  -1.4777589028      0
#> 326   0.4299453379      0
#> 327  -0.8135955278      0
#> 328   1.1643172309      0
#> 329   1.7959732151      0
#> 330  -0.4670780927      0
#> 331  -0.7724173750      0
#> 332   1.2212643583      0
#> 333  -1.9002153796      0
#> 334   0.8454908943      0
#> 335  -1.5153532639      0
#> 336   0.7479809658      0
#> 337   0.6760448775      0
#> 338  -0.2148621093      0
#> 339   1.4319383828      0
#> 340  -0.4374789259      0
#> 341  -0.7755987875      0
#> 342   1.1410848673      0
#> 343  -2.3953103016      0
#> 344   0.4866985326      0
#> 345   0.8862421469      0
#> 346   0.9022715393      0
#> 347   0.0679484043      0
#> 348   1.1688438190      0
#> 349  -0.5169322443      0
#> 350  -0.0381245098      0
#> 351   0.3898159459      0
#> 352   1.4621924094      0
#> 353  -1.8299151240      0
#> 354   1.0055930183      0
#> 355  -0.0270528929      0
#> 356   0.2036376214      0
#> 357   0.1602498202      0
#> 358   0.1671595160      0
#> 359   0.4548606474      0
#> 360  -1.3313986133      0
#> 361  -0.2608056809      0
#> 362  -1.3316222487      0
#> 363   1.0593971037      0
#> 364   2.5552422521      0
#> 365   1.9844459575      0
#> 366   2.0601869523      0
#> 367  -0.4609465988      0
#> 368  -0.1008370907      0
#> 369  -0.1275768354      0
#> 370  -0.5533204031      0
#> 371  -0.5884182064      0
#> 372   0.4648423714      0
#> 373  -0.0629054172      0
#> 374  -0.8151986882      0
#> 375   1.5090738062      0
#> 376   0.8933411346      0
#> 377   0.2384669699      0
#> 378   0.7581766134      0
#> 379  -1.0547333893      0
#> 380   1.1436344574      0
#> 381  -1.5472539974      0
#> 382  -0.1844172565      0
#> 383  -1.1501499212      0
#> 384  -0.2417184215      0
#> 385  -0.3642523730      0
#> 386   0.4333811759      0
#> 387  -0.2501802974      0
#> 388   1.2389756773      0
#> 389  -1.0980898544      0
#> 390   0.8542099994      0
#> 391   0.2835346888      0
#> 392   1.8640953155      0
#> 393  -0.3896257074      0
#> 394   0.9120265560      0
#> 395   0.2861572701      0
#> 396  -0.3555807683      0
#> 397   1.2780780474      0
#> 398   0.8588814652      0
#> 399   1.3701934572      0
#> 400  -1.2998467449      0
#> 401   1.0773268215      0
#> 402   0.4805307537      0
#> 403   0.2272849462      0
#> 404   0.1839061454      0
#> 405  -0.5151930456      0
#> 406  -1.1425338043      0
#> 407  -0.6634367707      0
#> 408  -0.1149208969      0
#> 409   0.7300580946      0
#> 410  -0.2499608331      0
#> 411   0.6083173666      0
#> 412   2.5855352041      0
#> 413   0.1367825832      0
#> 414  -1.2040209690      0
#> 415  -0.0642122379      0
#> 416  -0.0020405568      0
#> 417  -0.6215189788      0
#> 418  -1.5028715080      0
#> 419  -0.6649311658      0
#> 420   0.9889412216      0
#> 421   1.0016657997      0
#> 422   0.7400439480      0
#> 423   0.4278810907      0
#> 424  -0.8808494847      0
#> 425   0.6571593616      0
#> 426  -0.0652518514      0
#> 427  -1.2526761502      0
#> 428  -0.5389259982      0
#> 429   1.1316927501      0
#> 430  -2.1656470283      0
#> 431   2.6010157248      0
#> 432   1.0689181438      0
#> 433  -0.2937100124      0
#> 434   0.2165869035      0
#> 435   1.9265225209      0
#> 436   2.5518653382      0
#> 437  -1.3587825846      0
#> 438  -0.2958061404      0
#> 439   1.8728504678      0
#> 440   0.0970438550      0
#> 441   0.1960688249      0
#> 442  -0.7260796316      0
#> 443   0.4439146839      0
#> 444  -0.0315319799      0
#> 445   1.4743113558      0
#> 446  -0.4030722019      0
#> 447  -0.3806094384      0
#> 448   0.7112104731      0
#> 449  -1.4382778342      0
#> 450  -0.6758863000      0
#> 451  -0.4000498025      0
#> 452   1.1977101062      0
#> 453  -0.3617128499      0
#> 454   0.2659236596      0
#> 455   0.2763397152      0
#> 456   0.1657437470      0
#> 457   1.2286699588      0
#> 458   0.0116556308      0
#> 459  -0.9839113374      0
#> 460   0.1164566906      0
#> 461  -0.3631606226      0
#> 462   0.1252596441      0
#> 463   0.4222907792      0
#> 464   0.3753013241      0
#> 465  -1.2185336892      0
#> 466  -1.5397113298      0
#> 467   1.1758507351      0
#> 468   0.1081299243      0
#> 469   0.2877915123      0
#> 470   0.3279334038      0
#> 471   0.3019449458      0
#> 472   0.6220016167      0
#> 473  -0.9603802150      0
#> 474   1.2790177271      0
#> 475  -2.4472606652      0
#> 476  -0.7329908976      0
#> 477   1.5042205085      0
#> 478   0.1012896055      0
#> 479  -1.5669575599      0
#> 480   1.0618252484      0
#> 481   0.1075668849      0
#> 482  -0.6927877104      0
#> 483   2.7454215138      0
#> 484  -1.4633810863      0
#> 485  -0.6742143664      0
#> 486  -0.2126284009      0
#> 487   1.1563000990      0
#> 488   0.5176302067      0
#> 489  -0.1324464964      0
#> 490  -1.1040935945      0
#> 491   0.2950272949      0
#> 492  -0.3866609702      0
#> 493   1.1548603656      0
#> 494  -0.6221551953      0
#> 495  -1.3672508170      0
#> 496   0.7689621287      0
#> 497  -0.2500855405      0
#> 498  -0.5650540128      0
#> 499  -2.2038091364      0
#> 500   0.9929241214      0
#> 501  -0.2688536739      0
#> 502  -1.5111734653      0
#> 503   0.2650959139      0
#> 504   0.5227976161      0
#> 505   1.3389300517      0
#> 506  -0.4476190681      0
#> 507   0.3450312140      0
#> 508  -0.0573603921      0
#> 509  -0.4785174985      0
#> 510  -0.0258510145      0
#> 511   0.2562961817      0
#> 512  -2.5385103764      0
#> 513  -1.2027491012      0
#> 514  -0.2932451554      0
#> 515   1.4210253882      0
#> 516   0.6882001157      0
#> 517   0.2502846523      0
#> 518   0.7824111481      0
#> 519   0.3384963380      0
#> 520  -1.2713405055      0
#> 521   0.7923096665      0
#> 522   1.8779780687      0
#> 523   0.6471081633      0
#> 524   0.3626040430      0
#> 525  -1.1557238852      0
#> 526  -1.1039753253      0
#> 527  -0.8737434918      0
#> 528  -0.3815793110      0
#> 529  -0.2947733287      0
#> 530   1.3417654837      0
#> 531  -0.4082005315      0
#> 532   0.0158072135      0
#> 533   0.1675181823      0
#> 534   1.6619479108      0
#> 535  -0.2588017927      0
#> 536  -0.8689584370      0
#> 537   0.3355494033      0
#> 538   0.2811702007      0
#> 539   0.7033255915      0
#> 540  -0.9600206884      0
#> 541  -1.1756076635      0
#> 542  -0.3909027679      0
#> 543  -0.0364291512      0
#> 544  -1.8157113792      0
#> 545  -0.5694590525      0
#> 546   1.2194397002      0
#> 547   1.0623066655      0
#> 548   1.6663460769      0
#> 549   0.3452592954      0
#> 550  -0.3667540602      0
#> 551   1.7064386627      0
#> 552   0.1812932131      0
#> 553  -1.5861647316      0
#> 554   1.5611651215      0
#> 555   0.6154374306      0
#> 556   0.6565610367      0
#> 557  -0.3753445490      0
#> 558   0.4129507432      0
#> 559  -1.8017270106      0
#> 560   1.0945090368      0
#> 561   1.3986436719      0
#> 562  -0.8354878710      0
#> 563  -1.7306807905      0
#> 564  -0.4209941450      0
#> 565   0.5928777008      0
#> 566   0.7820327169      0
#> 567  -1.8557784005      0
#> 568   0.7250555538      0
#> 569   0.3865506459      0
#> 570   1.3461635493      0
#> 571  -0.3361304675      0
#> 572   2.3008349726      0
#> 573  -0.0454438328      0
#> 574   0.8616822063      0
#> 575  -1.1388029636      0
#> 576  -1.9913543547      0
#> 577   0.3833393067      0
#> 578  -0.4880425486      0
#> 579  -2.8081047740      0
#> 580   0.5697655043      0
#> 581   0.0048559716      0
#> 582   0.9809774707      0
#> 583  -0.0361190106      0
#> 584   0.4420633608      0
#> 585  -0.7620366496      0
#> 586   1.0144329327      0
#> 587  -1.0139557500      0
#> 588  -2.0281397704      0
#> 589  -1.1088928480      0
#> 590  -0.4791150014      0
#> 591   0.0553260420      0
#> 592   0.0576446319      0
#> 593  -1.7911485485      0
#> 594  -0.3303927592      0
#> 595  -3.2896767650      0
#> 596  -0.0906951171      0
#> 597  -0.9018452185      0
#> 598   1.9356952408      0
#> 599   0.4063034168      0
#> 600   0.7033877368      0
#> 601   0.1604215709      0
#> 602   2.2915127093      0
#> 603   0.1481596881      0
#> 604   1.7635893785      0
#> 605  -0.3374670121      0
#> 606   0.2949438400      0
#> 607   1.3428129265      0
#> 608  -0.4913969025      0
#> 609  -1.3096040423      0
#> 610  -0.2333645127      0
#> 611   0.8392383048      0
#> 612   1.4958334920      0
#> 613   0.6933088379      0
#> 614  -0.2602141035      0
#> 615  -1.8321013668      0
#> 616   0.4370724347      0
#> 617   0.4089215850      0
#> 618   0.3920730879      0
#> 619   1.7731553749      0
#> 620   0.2114481714      0
#> 621  -0.4841299330      0
#> 622  -0.5373389253      0
#> 623  -0.1436413889      0
#> 624  -0.7888093623      0
#> 625  -1.9973636284      0
#> 626   0.8312925662      0
#> 627   0.0495239301      0
#> 628  -1.7585262711      0
#> 629  -0.1852994257      0
#> 630   0.2679153636      0
#> 631   0.4385957034      0
#> 632  -0.6538781037      0
#> 633  -0.2024463038      0
#> 634   1.3141732940      0
#> 635   0.4026120650      0
#> 636  -0.8484019043      0
#> 637  -2.5828320897      0
#> 638   1.4730122418      0
#> 639  -0.6046358396      0
#> 640  -0.8863677916      0
#> 641  -0.1722436295      0
#> 642  -0.5736328183      0
#> 643   0.3639918451      0
#> 644   0.2150688100      0
#> 645  -0.9442058036      0
#> 646  -2.1703542533      0
#> 647  -0.3767975829      0
#> 648  -1.2413395100      0
#> 649   0.0512978201      0
#> 650  -1.0388643569      0
#> 651  -0.4868867849      0
#> 652   1.8602321258      0
#> 653  -0.1308874703      0
#> 654   0.7336912971      0
#> 655  -0.3437614721      0
#> 656  -1.0227941628      0
#> 657  -0.6187212252      0
#> 658   1.3074584637      0
#> 659   1.6249853327      0
#> 660   0.8808684756      0
#> 661  -0.1687108764      0
#> 662  -0.4881498854      0
#> 663  -0.9009304852      0
#> 664  -1.6769397831      0
#> 665  -0.3016372042      0
#> 666  -0.6712422469      0
#> 667   1.3052678827      0
#> 668   0.4751422693      0
#> 669  -1.5612159602      0
#> 670   1.2056503260      0
#> 671  -0.7894810107      0
#> 672  -0.7955352303      0
#> 673   0.6837303592      0
#> 674  -0.1650385538      0
#> 675   0.2562347371      0
#> 676  -0.2302449327      0
#> 677   0.5324375482      0
#> 678   1.2224311853      0
#> 679  -1.1750039479      0
#> 680  -1.0218644065      0
#> 681   0.6619675644      0
#> 682  -0.8877521520      0
#> 683  -2.3345208168      0
#> 684  -0.1261676828      0
#> 685   1.1762355238      0
#> 686   0.4349214962      0
#> 687  -1.4481262863      0
#> 688  -0.3091940979      0
#> 689  -0.9724609256      0
#> 690   0.2316211308      0
#> 691   0.3567585376      0
#> 692   1.5387693686      0
#> 693  -0.8573335106      0
#> 694   1.0701059572      0
#> 695   0.4554723696      0
#> 696  -0.0116987151      0
#> 697  -1.1105470799      0
#> 698  -0.7213114999      0
#> 699  -1.2610072615      0
#> 700  -1.7003253744      0
#> 701   0.7881856472      0
#> 702  -0.6901250798      0
#> 703  -1.4291544751      0
#> 704   1.4107096206      0
#> 705  -0.0241515713      0
#> 706  -0.5405695209      0
#> 707  -0.0550990207      0
#> 708  -0.1905142294      0
#> 709  -0.4091189642      0
#> 710  -0.0794572141      0
#> 711  -0.4113604301      0
#> 712   1.1767786537      0
#> 713  -0.5012794080      0
#> 714  -0.1321644426      0
#> 715  -0.7600493743      0
#> 716  -0.2462606595      0
#> 717   1.6546807666      0
#> 718  -0.1933859563      0
#> 719  -0.8469311238      0
#> 720   2.8811601725      0
#> 721  -0.8058524796      0
#> 722   0.2613728622      0
#> 723   1.8892196256      0
#> 724   0.0242095917      0
#> 725  -1.0175632548      0
#> 726  -0.3439296267      0
#> 727   1.0915620287      0
#> 728  -1.4282653981      0
#> 729  -0.8242659170      0
#> 730   1.3425974189      0
#> 731  -1.0030648211      0
#> 732   0.2397226285      0
#> 733   0.5825920445      0
#> 734   0.5789073250      0
#> 735  -0.3767181078      0
#> 736   0.9144568610      0
#> 737  -0.8657589052      0
#> 738  -1.0733972017      0
#> 739   0.3096796507      0
#> 740   0.7409234255      0
#> 741   0.4651022728      0
#> 742  -1.0964945405      0
#> 743   0.0455844008      0
#> 744   0.5476523905      0
#> 745  -0.3522634629      0
#> 746  -0.7103782592      0
#> 747  -0.3915058857      0
#> 748  -1.6193677648      0
#> 749  -0.4492082669      0
#> 750  -2.0693077369      0
#> 751  -0.6266102229      0
#> 752  -1.3050638738      0
#> 753  -0.5236527394      0
#> 754   0.0168525695      0
#> 755   1.0912902186      0
#> 756   0.2831393769      0
#> 757   0.1655262349      0
#> 758   0.8311173646      0
#> 759   0.1768087140      0
#> 760   0.4343196990      0
#> 761   1.6126414999      0
#> 762  -0.2534639575      0
#> 763   1.2147067096      0
#> 764  -0.5856682057      0
#> 765  -0.8744677245      0
#> 766  -0.0613625496      0
#> 767   0.2973938718      0
#> 768   1.3242888519      0
#> 769  -0.7193010218      0
#> 770   1.8172039294      0
#> 771   0.9147027115      0
#> 772  -0.3360259516      0
#> 773   0.9989246170      0
#> 774   0.4455596749      0
#> 775  -1.1576016079      0
#> 776   0.6055089071      0
#> 777   0.9878467333      0
#> 778  -0.6087020133      0
#> 779   0.0421188347      0
#> 780   0.6036264196      0
#> 781  -0.4932455735      0
#> 782   0.1604629790      0
#> 783   0.3669994336      0
#> 784   0.3641626194      0
#> 785  -0.7564448071      0
#> 786  -0.2164109535      0
#> 787  -0.1705731749      0
#> 788  -0.4935271901      0
#> 789   2.6568337026      0
#> 790   0.0470904634      0
#> 791  -1.0926186981      0
#> 792   2.1677150737      0
#> 793  -1.1594927681      0
#> 794   0.9471635983      0
#> 795  -0.0463608470      0
#> 796   0.1726590221      0
#> 797  -2.0734381974      0
#> 798   0.7765638282      0
#> 799   0.1898282472      0
#> 800  -1.8353922292      0
#> 801  -0.9685276214      0
#> 802   1.2556472640      0
#> 803   0.0887103537      0
#> 804  -0.4298531902      0
#> 805   2.3251403426      0
#> 806  -1.5084334631      0
#> 807  -0.0604230408      0
#> 808   0.5198259212      0
#> 809   1.4625179771      0
#> 810   0.3435039972      0
#> 811   0.9058168835      0
#> 812  -0.9333550768      0
#> 813   0.0719079094      0
#> 814   0.3637402937      0
#> 815   0.0959556626      0
#> 816  -1.2528989759      0
#> 817  -0.3066778216      0
#> 818   0.2186649666      0
#> 819  -0.0840406274      0
#> 820  -0.3654366682      0
#> 821   1.9682225479      0
#> 822  -0.2844195751      0
#> 823   0.0752566112      0
#> 824  -0.7699161588      0
#> 825  -0.6016366810      0
#> 826   0.4627608877      0
#> 827  -1.0158518621      0
#> 828  -0.2326804811      0
#> 829  -1.7223187809      0
#> 830  -2.2303992061      0
#> 831  -1.2886154804      0
#> 832   1.1298061814      0
#> 833  -0.2505853121      0
#> 834   0.5129743168      0
#> 835   0.6451903735      0
#> 836  -0.5046395357      0
#> 837   0.0302151013      0
#> 838   0.8545476494      0
#> 839  -0.5395052797      0
#> 840   1.3490329204      0
#> 841  -0.1098420470      0
#> 842  -0.6266019288      0
#> 843  -0.6750873395      0
#> 844   0.6500242230      0
#> 845   0.1371038348      0
#> 846  -1.1711282659      0
#> 847  -0.8976799565      0
#> 848  -0.5694575751      0
#> 849  -0.1209506420      0
#> 850  -0.8239260083      0
#> 851  -0.4460491502      0
#> 852  -1.4147705752      0
#> 853  -0.3066803939      0
#> 854  -0.6200988153      0
#> 855  -1.6906616180      0
#> 856  -0.5825695629      0
#> 857   0.1838969569      0
#> 858  -0.0766170088      0
#> 859   0.4358332243      0
#> 860  -0.1999629521      0
#> 861  -0.2932628314      0
#> 862  -0.5432094821      0
#> 863  -0.8013652377      0
#> 864   1.2665880727      0
#> 865   0.4570844600      0
#> 866   0.1098416712      0
#> 867   0.3859991310      0
#> 868   1.7340769181      0
#> 869  -0.4860962798      0
#> 870   0.7053302860      0
#> 871  -0.4258074576      0
#> 872   0.6681076215      0
#> 873   1.2850235034      0
#> 874   0.9537221609      0
#> 875   0.6553695909      0
#> 876   1.8331299945      0
#> 877  -2.5611054435      0
#> 878   0.6018847590      0
#> 879   0.3000915801      0
#> 880   2.0434613234      0
#> 881  -0.5382893132      0
#> 882   0.7231317792      0
#> 883   0.3083432432      0
#> 884  -0.8856379561      0
#> 885   1.3186809818      0
#> 886  -0.0410468657      0
#> 887  -1.4090044356      0
#> 888  -1.7689819104      0
#> 889  -1.0522093895      0
#> 890   0.0606039015      0
#> 891  -1.3383033030      0
#> 892   0.2171146159      0
#> 893   0.2397816000      0
#> 894  -2.0255330349      0
#> 895  -0.2465825211      0
#> 896  -0.8251417917      0
#> 897   0.6581989437      0
#> 898   0.6642605264      0
#> 899   1.3611408549      0
#> 900   0.7484134829      0
#> 901  -0.2166416102      0
#> 902   0.0804708179      0
#> 903  -1.7147048201      0
#> 904   2.1535766909      0
#> 905   0.6795311077      0
#> 906   1.2045665584      0
#> 907  -1.6467349064      0
#> 908   1.4830001512      0
#> 909   0.3342501818      1
#> 910   0.3934110285      1
#> 911   0.2926974090      1
#> 912   1.1982997176      1
#> 913   1.3469959684      1
#> 914   0.9677760393      1
#> 915   2.1566235100      1
#> 916   1.5675603413      1
#> 917   0.1716189821      1
#> 918  -0.2158213460      1
#> 919   0.4148758503      1
#> 920   1.1032983876      1
#> 921   1.3829733823      1
#> 922   0.7355232593      1
#> 923   2.0865886713      1
#> 924   1.9817696049      1
#> 925   2.6171166953      1
#> 926   2.2195413913      1
#> 927   1.1194750046      1
#> 928  -0.8177111530      1
#> 929   1.3503658995      1
#> 930   1.7737164465      1
#> 931   1.6525758287      1
#> 932   0.5486450190      1
#> 933   1.2350980953      1
#> 934   1.6492676410      1
#> 935   1.7677798007      1
#> 936   1.8955732646      1
#> 937   0.3218754736      1
#> 938   1.3294230195      1
#> 939  -0.4220151763      1
#> 940   1.1138070041      1
#> 941   1.6058200817      1
#> 942  -0.5601161447      1
#> 943   1.4641344990      1
#> 944   1.4717796420      1
#> 945   0.3412237437      1
#> 946   2.4279541050      1
#> 947  -0.4666387886      1
#> 948  -1.4252295542      1
#> 949   1.5216128568      1
#> 950   0.9077885922      1
#> 951   2.5457209049      1
#> 952   1.6106818725      1
#> 953  -1.0199833312      1
#> 954   1.2115782651      1
#> 955   1.2662882348      1
#> 956   2.4737765185      1
#> 957   1.1716268894      1
#> 958   1.3363420228      1
#> 959   0.6428204018      1
#> 960  -1.3685846324      1
#> 961   0.2550579058      1
#> 962   1.4721556334      1
#> 963   0.4462608458      1
#> 964   0.2032410787      1
#> 965   0.9667395767      1
#> 966   1.2854701176      1
#> 967  -0.5222901652      1
#> 968  -0.2973337378      1
#> 969   1.7118022994      1
#> 970  -0.5885550984      1
#> 971   1.6760666176      1
#> 972   0.1512405467      1
#> 973   1.4110334114      1
#> 974   0.8657925547      1
#> 975   1.7559174424      1
#> 976   0.7602005758      1
#> 977   1.6373988665      1
#> 978   0.5810434066      1
#> 979   0.9468860713      1
#> 980   0.6060124970      1
#> 981   0.1116820663      1
#> 982   1.2818994627      1
#> 983  -2.0174894820      1
#> 984   1.9134248962      1
#> 985  -0.4235322423      1
#> 986  -0.2054152993      1
#> 987   0.9296774804      1
#> 988  -0.1083761233      1
#> 989   0.8749778648      1
#> 990  -0.0734167759      1
#> 991   2.5938947077      1
#> 992   0.3513614957      1
#> 993   1.1653508913      1
#> 994   1.0494822836      1
#> 995  -0.7352412892      1
#> 996   0.8159179750      1
#> 997   1.0655742045      1
#> 998   0.5716844708      1
#> 999  -0.2415345438      1
#> 1000  0.5217642139      1