Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -3.625510e-01      0
#> 2    -2.944016e-01      0
#> 3    -2.782934e-01      0
#> 4     3.116651e-01      0
#> 5    -4.803800e-01      0
#> 6     2.525860e-01      0
#> 7     2.647487e-01      0
#> 8    -4.657664e-01      0
#> 9     2.272234e-01      0
#> 10    3.877153e-01      0
#> 11    7.286075e-01      0
#> 12    5.897649e-01      0
#> 13    8.363711e-01      0
#> 14    7.217748e-02      0
#> 15    1.758211e+00      0
#> 16    1.612977e+00      0
#> 17   -1.712857e-01      0
#> 18   -9.504773e-01      0
#> 19    7.304522e-01      0
#> 20    1.107984e-01      0
#> 21   -3.788684e-01      0
#> 22    2.135583e-01      0
#> 23   -1.452999e-02      0
#> 24   -7.105718e-01      0
#> 25   -1.059609e+00      0
#> 26   -2.816808e+00      0
#> 27   -2.273627e+00      0
#> 28    2.615381e+00      0
#> 29   -8.904899e-01      0
#> 30    1.352916e+00      0
#> 31    3.804013e-01      0
#> 32    1.105526e+00      0
#> 33    5.299780e-01      0
#> 34   -5.951199e-02      0
#> 35    8.824409e-02      0
#> 36    8.319437e-01      0
#> 37    1.404806e+00      0
#> 38   -8.980403e-01      0
#> 39   -6.734391e-02      0
#> 40   -4.321150e-01      0
#> 41   -4.461993e-01      0
#> 42    1.415762e+00      0
#> 43    8.062365e-01      0
#> 44    7.646385e-01      0
#> 45    2.256195e+00      0
#> 46   -3.118347e+00      0
#> 47    6.754055e-01      0
#> 48    1.036324e+00      0
#> 49   -1.359343e+00      0
#> 50    9.325033e-01      0
#> 51    4.694259e-01      0
#> 52    1.688387e+00      0
#> 53   -1.412540e-01      0
#> 54   -3.541387e-02      0
#> 55   -8.891522e-01      0
#> 56    1.454789e-01      0
#> 57    4.878006e-01      0
#> 58    1.410190e+00      0
#> 59    8.652515e-01      0
#> 60    1.773348e+00      0
#> 61    1.391497e+00      0
#> 62    2.049742e+00      0
#> 63   -9.341234e-02      0
#> 64   -9.514907e-01      0
#> 65    4.105800e-01      0
#> 66   -2.776742e-01      0
#> 67    1.705779e+00      0
#> 68    4.674158e-01      0
#> 69   -3.625264e-01      0
#> 70   -1.123650e+00      0
#> 71   -1.057240e+00      0
#> 72   -5.601031e-01      0
#> 73    5.070429e-01      0
#> 74    9.570833e-01      0
#> 75   -4.924270e-01      0
#> 76   -3.092039e-01      0
#> 77   -8.369128e-01      0
#> 78   -3.121264e-02      0
#> 79   -1.575637e+00      0
#> 80    3.200654e-01      0
#> 81   -1.062410e+00      0
#> 82   -3.951835e-01      0
#> 83    2.437335e-01      0
#> 84   -4.524212e-02      0
#> 85    8.110434e-01      0
#> 86   -2.642010e-01      0
#> 87    1.342009e+00      0
#> 88    1.176537e-01      0
#> 89    7.173899e-01      0
#> 90   -1.316269e+00      0
#> 91   -5.131414e-01      0
#> 92    8.102454e-01      0
#> 93   -5.810370e-01      0
#> 94    1.728779e+00      0
#> 95   -4.804264e-01      0
#> 96   -1.837499e+00      0
#> 97   -7.008697e-01      0
#> 98   -1.397030e-01      0
#> 99   -1.433379e-01      0
#> 100  -1.501777e-01      0
#> 101  -8.463407e-01      0
#> 102  -1.351369e+00      0
#> 103  -7.605192e-01      0
#> 104   4.898822e-01      0
#> 105  -9.172599e-01      0
#> 106   1.318242e-01      0
#> 107  -1.229534e+00      0
#> 108   7.762347e-01      0
#> 109  -6.280824e-01      0
#> 110  -9.292996e-01      0
#> 111   1.687009e-01      0
#> 112  -3.231611e-02      0
#> 113  -9.156208e-01      0
#> 114  -1.054957e-01      0
#> 115   1.939748e+00      0
#> 116   1.972475e-01      0
#> 117  -6.667725e-01      0
#> 118  -2.226978e-01      0
#> 119   4.691309e-01      0
#> 120  -3.452972e-01      0
#> 121  -2.265865e-01      0
#> 122   1.995971e+00      0
#> 123  -3.081038e-02      0
#> 124   1.089396e+00      0
#> 125   2.999858e+00      0
#> 126  -5.663091e-04      0
#> 127  -4.605130e-01      0
#> 128  -2.882101e-01      0
#> 129   1.521411e-01      0
#> 130   1.072422e+00      0
#> 131  -1.161113e+00      0
#> 132   8.167212e-01      0
#> 133   9.102432e-01      0
#> 134  -2.002113e-01      0
#> 135  -2.527645e+00      0
#> 136   6.297467e-01      0
#> 137  -1.623210e+00      0
#> 138  -7.602911e-01      0
#> 139  -2.382111e-01      0
#> 140   6.799201e-01      0
#> 141  -1.327639e-01      0
#> 142   1.067644e+00      0
#> 143   1.485190e-01      0
#> 144  -1.487370e+00      0
#> 145   1.210854e+00      0
#> 146   7.522669e-01      0
#> 147   7.127391e-01      0
#> 148  -7.549644e-01      0
#> 149   8.387118e-01      0
#> 150  -5.385326e-01      0
#> 151   5.970761e-01      0
#> 152  -5.366872e-01      0
#> 153  -3.940124e-01      0
#> 154  -8.060940e-01      0
#> 155  -1.550243e+00      0
#> 156  -5.028834e-01      0
#> 157  -1.804233e-01      0
#> 158  -7.429085e-01      0
#> 159  -1.729045e+00      0
#> 160  -1.050400e-01      0
#> 161  -1.279117e+00      0
#> 162  -9.339124e-01      0
#> 163  -1.776118e+00      0
#> 164   1.049750e+00      0
#> 165  -1.481824e+00      0
#> 166  -4.329507e-01      0
#> 167   9.598208e-01      0
#> 168  -3.091159e-01      0
#> 169   1.073689e+00      0
#> 170   8.048511e-01      0
#> 171  -4.777653e-01      0
#> 172  -9.364596e-01      0
#> 173  -5.480847e-01      0
#> 174   4.955762e-01      0
#> 175  -1.618901e-01      0
#> 176   4.467125e-01      0
#> 177   1.363900e+00      0
#> 178   5.144114e-01      0
#> 179   1.471816e+00      0
#> 180   1.121919e+00      0
#> 181   2.689775e-02      0
#> 182  -7.502653e-01      0
#> 183   1.628561e+00      0
#> 184   1.132656e+00      0
#> 185  -5.246300e-01      0
#> 186   1.102384e-01      0
#> 187  -3.036464e-01      0
#> 188   6.411190e-01      0
#> 189  -9.684940e-02      0
#> 190   7.713500e-01      0
#> 191   1.460584e+00      0
#> 192  -8.990241e-01      0
#> 193   2.631967e-01      0
#> 194   6.474377e-01      0
#> 195   3.996696e-01      0
#> 196   1.182948e+00      0
#> 197   1.123730e+00      0
#> 198   1.085378e+00      0
#> 199  -7.177916e-01      0
#> 200   8.736662e-01      0
#> 201  -8.224883e-01      0
#> 202   4.252336e-01      0
#> 203   2.127402e-01      0
#> 204   6.724159e-02      0
#> 205  -1.476124e-01      0
#> 206  -4.183383e-02      0
#> 207   2.539067e-01      0
#> 208  -1.683753e-01      0
#> 209   1.487672e-02      0
#> 210   2.093282e-02      0
#> 211   3.956739e-01      0
#> 212   5.472537e-01      0
#> 213  -1.777694e-01      0
#> 214   9.550059e-01      0
#> 215   9.605880e-01      0
#> 216  -1.858405e-01      0
#> 217   1.551498e+00      0
#> 218   8.259774e-01      0
#> 219   9.420488e-01      0
#> 220  -2.084864e+00      0
#> 221   4.489843e-01      0
#> 222  -4.157961e-01      0
#> 223   5.723967e-03      0
#> 224   6.259841e-01      0
#> 225  -1.180066e-01      0
#> 226  -6.744430e-01      0
#> 227   1.948984e+00      0
#> 228   4.873955e-01      0
#> 229   6.321129e-02      0
#> 230   9.518736e-01      0
#> 231  -1.246647e+00      0
#> 232  -3.078060e-01      0
#> 233  -1.005589e+00      0
#> 234   7.960629e-01      0
#> 235  -2.210554e+00      0
#> 236   2.783378e-01      0
#> 237   3.829865e-01      0
#> 238   2.052616e+00      0
#> 239   1.110738e+00      0
#> 240  -4.294458e-01      0
#> 241   2.583978e-01      0
#> 242  -4.803846e-01      0
#> 243  -1.596589e+00      0
#> 244   8.397496e-01      0
#> 245   5.787249e-03      0
#> 246   5.460601e-01      0
#> 247   1.431989e-01      0
#> 248  -1.406073e-01      0
#> 249   1.070334e+00      0
#> 250   9.458529e-01      0
#> 251  -2.642595e-01      0
#> 252  -1.044908e-01      0
#> 253  -1.971574e-01      0
#> 254   4.961616e-01      0
#> 255   9.094159e-02      0
#> 256   2.107524e+00      0
#> 257  -7.625605e-01      0
#> 258   7.366247e-01      0
#> 259  -4.344735e-01      0
#> 260  -6.217627e-01      0
#> 261   4.906544e-01      0
#> 262  -1.331852e-01      0
#> 263  -5.079280e-01      0
#> 264   7.791245e-02      0
#> 265  -7.958939e-01      0
#> 266   1.197933e+00      0
#> 267  -5.612101e-01      0
#> 268   7.911486e-02      0
#> 269  -1.985374e+00      0
#> 270   4.104236e-01      0
#> 271   3.506422e-01      0
#> 272  -1.075393e-01      0
#> 273  -1.795533e+00      0
#> 274   4.650242e-01      0
#> 275  -3.961234e-01      0
#> 276  -9.564651e-02      0
#> 277  -1.954258e-01      0
#> 278   4.748254e-01      0
#> 279   1.400392e+00      0
#> 280   8.440293e-01      0
#> 281   9.986702e-01      0
#> 282   9.099841e-01      0
#> 283   6.051031e-01      0
#> 284  -5.405861e-01      0
#> 285   2.725495e-01      0
#> 286   8.106890e-01      0
#> 287   2.111985e-01      0
#> 288   1.021510e+00      0
#> 289   1.909344e+00      0
#> 290  -5.618340e-01      0
#> 291   9.928774e-01      0
#> 292  -9.130042e-01      0
#> 293   7.351874e-01      0
#> 294  -8.467631e-02      0
#> 295  -2.262480e-02      0
#> 296  -1.887387e+00      0
#> 297   1.090338e+00      0
#> 298   3.898072e-01      0
#> 299   3.019833e-01      0
#> 300  -7.679564e-01      0
#> 301   1.496216e-01      0
#> 302   2.293464e-01      0
#> 303  -3.036234e-01      0
#> 304  -5.601852e-01      0
#> 305   1.294165e+00      0
#> 306   5.194279e-01      0
#> 307  -4.315842e-01      0
#> 308  -1.228725e-01      0
#> 309   1.585865e+00      0
#> 310   6.783376e-01      0
#> 311   6.350891e-01      0
#> 312   4.936046e-01      0
#> 313   5.518566e-01      0
#> 314   4.602286e-01      0
#> 315  -1.500131e-01      0
#> 316  -1.191652e+00      0
#> 317  -6.971826e-01      0
#> 318  -6.585791e-01      0
#> 319   2.453187e-01      0
#> 320  -1.533081e+00      0
#> 321   6.520569e-01      0
#> 322  -6.466077e-01      0
#> 323   2.932135e-01      0
#> 324  -1.302563e+00      0
#> 325   8.962707e-01      0
#> 326  -2.320973e-01      0
#> 327   7.354462e-02      0
#> 328   2.007285e+00      0
#> 329  -4.666827e-01      0
#> 330  -9.361820e-01      0
#> 331   3.763005e-01      0
#> 332  -5.433959e-01      0
#> 333   5.188379e-01      0
#> 334   5.977347e-01      0
#> 335  -2.039369e-01      0
#> 336   7.921793e-01      0
#> 337   2.876366e-01      0
#> 338   1.531475e+00      0
#> 339  -8.549333e-01      0
#> 340   3.207415e-01      0
#> 341  -4.894203e-01      0
#> 342  -1.379206e+00      0
#> 343  -3.922341e-01      0
#> 344   4.756003e-01      0
#> 345  -3.152815e-01      0
#> 346  -6.774725e-01      0
#> 347   2.083698e-01      0
#> 348  -7.322233e-01      0
#> 349   1.536497e-01      0
#> 350   7.495190e-01      0
#> 351   6.859697e-01      0
#> 352  -2.038284e+00      0
#> 353  -5.759766e-01      0
#> 354  -1.070798e-01      0
#> 355   3.222951e-01      0
#> 356   4.877438e-02      0
#> 357  -5.317043e-01      0
#> 358   3.735260e-01      0
#> 359   9.100446e-02      0
#> 360  -3.116298e-01      0
#> 361   7.571157e-01      0
#> 362   7.915833e-02      0
#> 363   2.795581e-01      0
#> 364   3.253208e-01      0
#> 365   3.977602e-01      0
#> 366   1.022809e+00      0
#> 367  -7.471474e-01      0
#> 368  -1.674413e+00      0
#> 369  -3.487463e-01      0
#> 370  -5.933024e-01      0
#> 371   8.986859e-01      0
#> 372   9.183826e-01      0
#> 373   1.370744e+00      0
#> 374  -7.438181e-02      0
#> 375   1.795737e+00      0
#> 376  -1.209077e+00      0
#> 377  -1.111196e+00      0
#> 378  -4.263336e-01      0
#> 379   5.649223e-01      0
#> 380  -1.239573e-01      0
#> 381  -1.227845e+00      0
#> 382  -1.708579e+00      0
#> 383   1.445738e+00      0
#> 384  -8.196765e-02      0
#> 385   9.112417e-01      0
#> 386  -2.266802e+00      0
#> 387   1.466935e+00      0
#> 388  -4.461715e-01      0
#> 389   1.193014e+00      0
#> 390  -8.416421e-01      0
#> 391  -3.821265e-01      0
#> 392   1.274900e+00      0
#> 393  -8.629057e-02      0
#> 394   4.975353e-01      0
#> 395  -2.194157e-01      0
#> 396  -8.377532e-01      0
#> 397   4.649889e-02      0
#> 398  -5.772652e-01      0
#> 399  -7.410452e-01      0
#> 400   7.384406e-01      0
#> 401   2.202471e+00      0
#> 402  -7.026946e-01      0
#> 403  -7.811461e-03      0
#> 404   2.506866e-01      0
#> 405  -4.551624e-01      0
#> 406  -1.217416e-01      0
#> 407   9.706990e-01      0
#> 408  -4.170734e-01      0
#> 409   1.161012e+00      0
#> 410  -1.134822e+00      0
#> 411  -1.064423e-01      0
#> 412   7.416303e-01      0
#> 413   1.188012e+00      0
#> 414   1.138487e-01      0
#> 415   1.681171e+00      0
#> 416  -7.711888e-01      0
#> 417   1.040064e+00      0
#> 418   4.361636e-01      0
#> 419  -9.418745e-01      0
#> 420   4.011723e-01      0
#> 421   7.846523e-02      0
#> 422  -1.695708e+00      0
#> 423  -4.764946e-01      0
#> 424  -2.181275e-02      0
#> 425   8.034134e-01      0
#> 426  -1.104750e-01      0
#> 427  -5.088747e-01      0
#> 428   1.705720e+00      0
#> 429   6.370239e-02      0
#> 430  -1.223612e+00      0
#> 431   1.245140e+00      0
#> 432   4.805407e-01      0
#> 433  -1.654170e-01      0
#> 434  -1.307791e+00      0
#> 435   6.426925e-02      0
#> 436  -2.326537e-01      0
#> 437  -4.918127e-01      0
#> 438   7.303465e-01      0
#> 439  -9.859434e-01      0
#> 440   8.947412e-01      0
#> 441  -2.001886e+00      0
#> 442  -3.112134e-01      0
#> 443  -1.101229e+00      0
#> 444   3.933232e-01      0
#> 445  -1.166945e+00      0
#> 446   1.364106e+00      0
#> 447  -1.054336e+00      0
#> 448  -6.233462e-01      0
#> 449   1.685134e+00      0
#> 450  -5.006965e-01      0
#> 451   4.927764e-02      0
#> 452   7.932745e-01      0
#> 453  -5.214269e-01      0
#> 454   5.261553e-01      0
#> 455  -1.211733e+00      0
#> 456   1.212972e+00      0
#> 457  -1.442237e+00      0
#> 458   8.744584e-01      0
#> 459   5.709298e-01      0
#> 460  -5.581162e-01      0
#> 461  -5.368399e-01      0
#> 462   1.541337e+00      0
#> 463  -2.030262e+00      0
#> 464   8.798392e-01      0
#> 465   1.089418e+00      0
#> 466  -7.190212e-01      0
#> 467  -1.504007e+00      0
#> 468  -9.271851e-02      0
#> 469   1.661584e-01      0
#> 470   8.686283e-01      0
#> 471   9.953944e-01      0
#> 472  -1.627100e-01      0
#> 473   7.276082e-02      0
#> 474  -1.303616e+00      0
#> 475  -1.401697e-01      0
#> 476  -8.187333e-02      0
#> 477   1.781375e+00      0
#> 478  -3.595059e-01      0
#> 479   1.595593e+00      0
#> 480  -1.571073e+00      0
#> 481   4.562419e-01      0
#> 482   3.749739e-01      0
#> 483   7.866212e-01      0
#> 484  -9.214224e-01      0
#> 485  -6.344170e-01      0
#> 486   6.477096e-01      0
#> 487   4.850644e-01      0
#> 488   4.226762e-01      0
#> 489   2.867234e-02      0
#> 490   2.409556e-02      0
#> 491   5.608042e-02      0
#> 492  -2.184050e+00      0
#> 493  -1.143448e+00      0
#> 494  -2.748861e-02      0
#> 495  -9.401404e-01      0
#> 496   1.465352e-01      0
#> 497   1.458332e-01      0
#> 498   1.413864e+00      0
#> 499   2.745001e-02      0
#> 500   1.361304e+00      0
#> 501  -1.105694e-01      0
#> 502   4.298759e-01      0
#> 503   1.427781e-01      0
#> 504  -1.885608e-01      0
#> 505  -2.222479e+00      0
#> 506  -7.696476e-01      0
#> 507   1.069783e+00      0
#> 508  -1.261614e+00      0
#> 509  -1.954259e+00      0
#> 510   5.570886e-01      0
#> 511   8.192956e-01      0
#> 512   1.923683e+00      0
#> 513  -1.756631e+00      0
#> 514   1.228699e+00      0
#> 515  -1.544456e-01      0
#> 516   6.182298e-01      0
#> 517   1.426541e-01      0
#> 518  -9.749516e-01      0
#> 519   1.226556e+00      0
#> 520  -4.386222e-01      0
#> 521   3.683266e-01      0
#> 522   1.442111e+00      0
#> 523   3.428162e-01      0
#> 524  -1.398338e+00      0
#> 525  -5.220216e-02      0
#> 526   1.270688e+00      0
#> 527   7.847260e-01      0
#> 528   1.895183e+00      0
#> 529  -9.960819e-01      0
#> 530  -3.739710e-01      0
#> 531   9.412379e-01      0
#> 532   1.973805e+00      0
#> 533  -8.956057e-01      0
#> 534  -1.912158e+00      0
#> 535   6.206250e-01      0
#> 536   1.812055e-01      0
#> 537  -1.092029e+00      0
#> 538   1.936196e-01      0
#> 539  -1.723889e+00      0
#> 540   1.004002e+00      0
#> 541   9.583966e-01      0
#> 542  -9.472686e-01      0
#> 543  -2.624855e+00      0
#> 544   9.219018e-01      0
#> 545  -2.305938e+00      0
#> 546   2.533470e-02      0
#> 547   2.204239e-02      0
#> 548  -9.845137e-01      0
#> 549  -1.691511e+00      0
#> 550  -9.192887e-01      0
#> 551   7.743649e-01      0
#> 552  -7.089911e-01      0
#> 553  -1.325293e+00      0
#> 554  -1.052741e+00      0
#> 555   1.457064e-01      0
#> 556   5.090219e-01      0
#> 557   1.085376e+00      0
#> 558   1.869157e-01      0
#> 559   3.176003e-01      0
#> 560   5.689727e-01      0
#> 561   1.606023e+00      0
#> 562   4.944744e-02      0
#> 563  -1.298057e-01      0
#> 564  -1.138451e+00      0
#> 565  -1.302585e+00      0
#> 566   1.194770e+00      0
#> 567  -4.489429e-01      0
#> 568  -4.540033e-01      0
#> 569   1.268298e+00      0
#> 570  -1.107435e+00      0
#> 571   1.157284e+00      0
#> 572  -1.086939e-01      0
#> 573  -2.592270e-01      0
#> 574  -1.523682e+00      0
#> 575   1.371310e+00      0
#> 576  -4.846713e-01      0
#> 577   4.769029e-01      0
#> 578  -4.977733e-02      0
#> 579  -5.457018e-01      0
#> 580  -6.399468e-01      0
#> 581   1.576785e+00      0
#> 582   4.547551e-01      0
#> 583  -8.385524e-01      0
#> 584  -4.007965e-01      0
#> 585  -1.177200e+00      0
#> 586  -6.922635e-01      0
#> 587  -3.905265e-01      0
#> 588  -1.801155e-01      0
#> 589   5.013339e-01      0
#> 590   6.848267e-02      0
#> 591   2.267387e+00      0
#> 592   8.303485e-01      0
#> 593  -1.404640e-01      0
#> 594   8.036856e-01      0
#> 595   1.021778e-01      0
#> 596   5.516391e-01      0
#> 597  -2.490552e+00      0
#> 598  -1.173210e+00      0
#> 599  -1.692338e+00      0
#> 600  -1.102544e+00      0
#> 601  -1.016595e+00      0
#> 602  -5.190578e-01      0
#> 603   2.485997e-01      0
#> 604  -1.719381e+00      0
#> 605  -2.091810e-01      0
#> 606   2.402358e-01      0
#> 607   2.711483e-01      0
#> 608   7.099598e-01      0
#> 609   9.319944e-01      0
#> 610   3.053439e-01      0
#> 611   7.017113e-01      0
#> 612   1.682657e+00      0
#> 613  -1.420005e+00      0
#> 614  -1.595292e+00      0
#> 615  -2.244684e-01      0
#> 616   6.906112e-01      0
#> 617  -4.757158e-01      0
#> 618   7.871579e-01      0
#> 619  -1.633495e+00      0
#> 620   1.388908e+00      0
#> 621   1.081042e+00      0
#> 622  -6.325546e-01      0
#> 623  -3.483468e-01      0
#> 624   1.025247e+00      0
#> 625   1.866550e+00      0
#> 626  -6.294682e-01      0
#> 627  -5.272756e-01      0
#> 628   3.525542e-01      0
#> 629  -8.766533e-02      0
#> 630  -9.667934e-01      0
#> 631   2.151019e-01      0
#> 632   4.712806e-01      0
#> 633  -1.297824e-01      0
#> 634  -1.053394e+00      0
#> 635   1.433627e+00      0
#> 636   1.587099e-01      0
#> 637   1.116481e+00      0
#> 638   2.667956e-01      0
#> 639  -1.416390e+00      0
#> 640  -5.533638e-02      0
#> 641  -1.339005e-01      0
#> 642  -1.949603e-01      0
#> 643  -1.191648e+00      0
#> 644   2.533406e-01      0
#> 645   1.037835e+00      0
#> 646   4.406865e-02      0
#> 647  -2.197926e+00      0
#> 648  -6.497191e-01      0
#> 649  -1.684867e+00      0
#> 650   8.885885e-01      0
#> 651   8.408950e-01      0
#> 652   5.183674e-01      0
#> 653   3.538828e-01      0
#> 654  -1.847149e+00      0
#> 655   2.219596e-01      0
#> 656   6.432391e-01      0
#> 657   3.982980e-01      0
#> 658   7.002169e-01      0
#> 659  -4.130726e-01      0
#> 660  -1.550777e+00      0
#> 661   7.381224e-01      0
#> 662  -1.327808e+00      0
#> 663  -1.395249e-01      0
#> 664  -9.205273e-01      0
#> 665   2.199894e-01      0
#> 666   1.486682e+00      0
#> 667  -2.955323e-02      0
#> 668   1.118495e+00      0
#> 669   2.344299e-01      0
#> 670   1.300735e+00      0
#> 671  -1.717670e+00      0
#> 672  -4.565065e-01      0
#> 673  -2.282455e-01      0
#> 674  -5.637160e-01      0
#> 675   2.441440e-01      0
#> 676   6.007698e-01      0
#> 677  -3.881390e-01      0
#> 678  -1.830478e-01      0
#> 679  -6.817738e-01      0
#> 680   1.385988e+00      0
#> 681  -7.706364e-01      0
#> 682  -2.057766e-01      0
#> 683  -7.727581e-01      0
#> 684   1.748964e-01      0
#> 685   1.205934e+00      0
#> 686  -3.415374e-01      0
#> 687   4.159619e-01      0
#> 688   8.272379e-01      0
#> 689   7.701334e-01      0
#> 690   6.102113e-01      0
#> 691   2.043703e+00      0
#> 692   6.370488e-01      0
#> 693   6.212768e-01      0
#> 694   1.875094e+00      0
#> 695  -5.778876e-01      0
#> 696   5.016987e-02      0
#> 697  -1.615463e+00      0
#> 698   7.865518e-01      0
#> 699  -6.044660e-01      0
#> 700   6.355193e-01      0
#> 701   2.314265e-02      0
#> 702   3.953843e-01      0
#> 703  -6.699577e-01      0
#> 704   3.665027e-01      0
#> 705   1.168345e+00      0
#> 706  -1.746534e-01      0
#> 707   1.551355e-01      0
#> 708   1.561314e+00      0
#> 709  -1.608509e+00      0
#> 710   1.204961e+00      0
#> 711  -1.717077e+00      0
#> 712   1.433500e+00      0
#> 713   4.090450e-01      0
#> 714   5.233221e-01      0
#> 715   2.926361e+00      0
#> 716   5.325595e-01      0
#> 717   2.190158e+00      0
#> 718   1.353968e+00      0
#> 719  -1.632346e+00      0
#> 720  -7.959198e-02      0
#> 721  -1.322057e+00      0
#> 722  -7.094756e-01      0
#> 723   9.749315e-01      0
#> 724   2.122066e+00      0
#> 725  -4.193582e-01      0
#> 726   4.249782e-01      0
#> 727   1.368330e+00      0
#> 728  -3.217308e-01      0
#> 729  -4.379006e-01      0
#> 730  -4.912212e-01      0
#> 731   1.911603e+00      0
#> 732   3.337507e-01      0
#> 733   7.151171e-01      0
#> 734   1.363749e-01      0
#> 735   4.504641e-01      0
#> 736   2.382425e-01      0
#> 737   1.817552e+00      0
#> 738  -5.197950e-01      0
#> 739   2.237802e+00      0
#> 740  -1.506746e+00      0
#> 741   7.491371e-02      0
#> 742   3.391996e-01      0
#> 743  -2.609824e-01      0
#> 744   1.698456e-01      0
#> 745   1.642589e+00      0
#> 746   2.555276e-01      0
#> 747  -1.970705e+00      0
#> 748   9.478054e-01      0
#> 749  -4.167972e-01      0
#> 750  -2.014933e-01      0
#> 751  -8.059370e-02      0
#> 752  -7.624546e-01      0
#> 753   2.351126e-01      0
#> 754   2.211597e+00      0
#> 755   7.357460e-01      0
#> 756   1.396544e+00      0
#> 757   3.763727e-01      0
#> 758   9.993556e-02      0
#> 759   1.555672e-01      0
#> 760  -3.882653e-02      0
#> 761  -3.430739e-01      0
#> 762  -4.557817e-01      0
#> 763   7.399803e-02      0
#> 764   1.054294e+00      0
#> 765   4.167582e-01      0
#> 766   1.189261e+00      0
#> 767  -1.240470e+00      0
#> 768   1.054137e+00      0
#> 769   1.710802e-01      0
#> 770   7.498451e-01      0
#> 771   2.800271e-01      0
#> 772   1.250464e+00      0
#> 773   1.003399e+00      0
#> 774   2.078407e-01      0
#> 775   1.910801e+00      0
#> 776  -1.213025e+00      0
#> 777   6.507016e-01      0
#> 778  -4.591558e-01      0
#> 779   9.605975e-01      0
#> 780   7.046941e-01      0
#> 781   1.714936e+00      0
#> 782  -1.034725e+00      0
#> 783   1.663391e-01      0
#> 784   7.822888e-01      0
#> 785   6.792556e-02      0
#> 786  -2.233616e-01      0
#> 787   9.719834e-01      0
#> 788   1.676381e+00      0
#> 789  -5.326628e-01      0
#> 790   1.091885e+00      0
#> 791   1.940139e+00      0
#> 792  -1.173600e+00      0
#> 793  -4.545530e-01      0
#> 794  -1.159892e+00      0
#> 795  -4.511501e-01      0
#> 796   1.393329e+00      0
#> 797  -4.593967e-01      0
#> 798   1.469521e+00      0
#> 799  -3.431384e-01      0
#> 800   7.734291e-01      0
#> 801  -8.253005e-01      0
#> 802   1.098347e+00      0
#> 803   1.405003e+00      0
#> 804  -7.131922e-01      0
#> 805  -1.797511e-01      0
#> 806   1.328937e+00      0
#> 807  -6.865859e-01      0
#> 808   6.794546e-01      0
#> 809  -4.791619e-01      0
#> 810  -4.764945e-01      0
#> 811  -2.646701e+00      0
#> 812  -2.896269e-01      0
#> 813  -5.042493e-01      0
#> 814   9.009309e-01      0
#> 815   7.580064e-02      0
#> 816   3.282247e-01      0
#> 817   1.698204e-01      0
#> 818  -8.898651e-02      0
#> 819   6.715312e-01      0
#> 820  -2.209846e+00      0
#> 821  -1.245286e+00      0
#> 822   9.211191e-01      0
#> 823   1.334186e+00      0
#> 824  -2.208586e-01      0
#> 825   8.495014e-01      0
#> 826  -1.529580e+00      0
#> 827  -1.931929e+00      0
#> 828   1.781639e+00      0
#> 829   1.749538e-02      0
#> 830   5.614803e-02      0
#> 831  -2.145065e+00      0
#> 832   9.599033e-01      0
#> 833   1.566308e+00      0
#> 834  -5.951642e-01      0
#> 835   4.855641e-01      0
#> 836  -1.989967e-01      0
#> 837  -1.885003e-01      0
#> 838   3.534608e-01      0
#> 839  -4.855152e-02      0
#> 840   1.771877e+00      0
#> 841  -7.308191e-01      0
#> 842  -2.490765e-01      0
#> 843   3.293067e-01      0
#> 844  -6.355674e-01      0
#> 845  -2.220530e-03      0
#> 846  -6.963851e-01      0
#> 847  -3.070263e-01      0
#> 848   2.289092e+00      0
#> 849   5.261281e-01      0
#> 850  -7.135129e-02      0
#> 851  -1.563417e+00      0
#> 852   2.581343e-01      0
#> 853  -1.635013e+00      0
#> 854   8.841578e-02      0
#> 855   1.860782e+00      0
#> 856   8.033326e-01      0
#> 857   8.783063e-02      0
#> 858   2.964893e-01      0
#> 859  -8.228942e-02      0
#> 860   5.497344e-01      0
#> 861  -4.823912e-01      0
#> 862  -3.146322e-01      0
#> 863  -1.307787e+00      0
#> 864  -1.129559e-01      0
#> 865   7.534137e-01      0
#> 866   3.070647e-01      0
#> 867  -1.119958e+00      0
#> 868   1.807138e-01      0
#> 869   3.127621e-01      0
#> 870  -2.449958e-01      0
#> 871   8.972690e-01      0
#> 872   1.940218e+00      0
#> 873   8.661708e-01      0
#> 874  -3.775089e-01      0
#> 875  -8.462790e-01      0
#> 876  -7.529757e-02      0
#> 877   1.021676e+00      0
#> 878   8.952478e-01      0
#> 879  -1.691318e-01      0
#> 880  -8.920150e-01      0
#> 881   2.182858e-01      0
#> 882   1.226231e+00      0
#> 883  -8.557468e-01      0
#> 884  -2.266945e-01      0
#> 885  -2.348720e-01      0
#> 886   1.109757e+00      0
#> 887  -7.461497e-01      0
#> 888   4.661644e-01      0
#> 889   3.813844e-01      0
#> 890   9.531241e-01      0
#> 891   7.526005e-01      0
#> 892   7.976176e-01      0
#> 893   7.228347e-01      0
#> 894  -2.490350e-01      0
#> 895   1.162589e+00      0
#> 896   1.659381e-01      0
#> 897  -1.828958e-01      1
#> 898   1.913109e+00      1
#> 899   1.319623e-01      1
#> 900  -8.758530e-02      1
#> 901   1.774723e+00      1
#> 902  -4.930655e-01      1
#> 903  -8.727605e-02      1
#> 904  -2.998759e-01      1
#> 905   1.115296e+00      1
#> 906   7.991167e-01      1
#> 907   1.510319e+00      1
#> 908   2.661823e+00      1
#> 909   7.654779e-01      1
#> 910   9.985193e-01      1
#> 911   1.041680e+00      1
#> 912  -4.199983e-01      1
#> 913  -1.846273e-01      1
#> 914   1.666223e+00      1
#> 915   1.448077e-01      1
#> 916   4.155994e-01      1
#> 917   1.059783e+00      1
#> 918   2.016120e-01      1
#> 919  -2.247692e+00      1
#> 920   1.445108e+00      1
#> 921   8.935322e-01      1
#> 922   1.118421e+00      1
#> 923  -4.472849e-02      1
#> 924   2.151711e+00      1
#> 925   1.209328e+00      1
#> 926   1.660754e-01      1
#> 927   1.400155e-01      1
#> 928   1.269056e-01      1
#> 929  -1.170791e+00      1
#> 930  -1.825840e-01      1
#> 931   4.773764e-01      1
#> 932   8.596136e-01      1
#> 933   9.726825e-01      1
#> 934   1.275084e+00      1
#> 935   8.426975e-01      1
#> 936  -1.477166e+00      1
#> 937   8.198401e-01      1
#> 938   5.832128e-01      1
#> 939   9.160649e-01      1
#> 940   7.853561e-01      1
#> 941   1.358820e+00      1
#> 942   8.295370e-01      1
#> 943   1.219640e+00      1
#> 944   1.209510e+00      1
#> 945  -1.403199e+00      1
#> 946   1.183627e+00      1
#> 947   6.981193e-01      1
#> 948   6.399265e-01      1
#> 949   2.755634e-01      1
#> 950  -6.239484e-01      1
#> 951   1.472342e+00      1
#> 952   1.135930e-01      1
#> 953  -5.877217e-01      1
#> 954   1.162243e+00      1
#> 955   1.155116e+00      1
#> 956   8.416508e-02      1
#> 957   1.390923e+00      1
#> 958  -6.317562e-01      1
#> 959   5.443780e-01      1
#> 960  -2.955775e-01      1
#> 961  -7.527073e-01      1
#> 962   1.097988e-01      1
#> 963   1.343747e+00      1
#> 964  -5.429863e-01      1
#> 965   2.276531e+00      1
#> 966  -1.176162e+00      1
#> 967  -8.464224e-01      1
#> 968   1.318306e+00      1
#> 969   1.130127e+00      1
#> 970   1.165338e+00      1
#> 971   5.738713e-01      1
#> 972   2.098126e-01      1
#> 973  -9.315930e-01      1
#> 974   4.593657e-01      1
#> 975   3.559941e-01      1
#> 976  -7.861207e-05      1
#> 977   1.130661e+00      1
#> 978   3.880016e-01      1
#> 979   1.836719e+00      1
#> 980  -2.926246e-02      1
#> 981   1.752228e+00      1
#> 982   3.325685e-01      1
#> 983  -7.127013e-01      1
#> 984  -9.436876e-02      1
#> 985   2.005866e+00      1
#> 986   1.013220e+00      1
#> 987   1.127493e+00      1
#> 988   1.871857e+00      1
#> 989   1.350172e+00      1
#> 990  -6.933104e-02      1
#> 991   7.492449e-01      1
#> 992   1.525295e+00      1
#> 993  -5.004140e-01      1
#> 994   1.462257e-02      1
#> 995   1.257278e+00      1
#> 996   1.219332e-01      1
#> 997   6.916678e-01      1
#> 998   8.675570e-01      1
#> 999   1.399705e-02      1
#> 1000  7.199147e-01      1