Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     3.485922e-01      0
#> 2    -1.790746e+00      0
#> 3    -1.094853e+00      0
#> 4     8.097290e-03      0
#> 5     5.763284e-01      0
#> 6     9.533116e-01      0
#> 7     5.625287e-01      0
#> 8     5.337399e-01      0
#> 9     1.023706e+00      0
#> 10    7.280076e-01      0
#> 11   -1.629165e+00      0
#> 12   -9.465085e-01      0
#> 13    5.509588e-01      0
#> 14    1.662195e-01      0
#> 15   -6.320436e-02      0
#> 16    7.156652e-01      0
#> 17   -5.220515e-01      0
#> 18   -5.082959e-01      0
#> 19    1.386304e+00      0
#> 20   -2.502150e-01      0
#> 21   -9.087300e-01      0
#> 22   -6.912892e-01      0
#> 23   -1.646937e+00      0
#> 24    9.726112e-02      0
#> 25    1.718564e+00      0
#> 26   -6.635379e-01      0
#> 27   -1.768955e-01      0
#> 28   -1.537058e+00      0
#> 29   -1.911500e+00      0
#> 30   -1.139235e+00      0
#> 31   -7.582907e-01      0
#> 32    5.330528e-01      0
#> 33   -3.049451e-01      0
#> 34    6.060548e-01      0
#> 35   -7.216790e-01      0
#> 36    9.549331e-01      0
#> 37    1.003005e-01      0
#> 38   -8.653348e-01      0
#> 39    1.100681e+00      0
#> 40   -4.206339e-01      0
#> 41    8.620616e-01      0
#> 42   -1.016252e-01      0
#> 43   -1.299079e-01      0
#> 44   -6.278482e-01      0
#> 45   -2.771784e-01      0
#> 46    3.208311e-01      0
#> 47   -1.314625e+00      0
#> 48    1.946140e+00      0
#> 49   -8.984778e-02      0
#> 50   -8.489787e-02      0
#> 51    6.820265e-01      0
#> 52   -1.229114e+00      0
#> 53   -1.150274e-01      0
#> 54    1.835082e-01      0
#> 55   -1.541635e+00      0
#> 56    1.564897e+00      0
#> 57   -1.392824e-01      0
#> 58   -7.503914e-01      0
#> 59    1.250553e-02      0
#> 60    4.021220e-01      0
#> 61   -2.335094e-01      0
#> 62   -1.570073e+00      0
#> 63   -1.965512e-01      0
#> 64   -6.478459e-01      0
#> 65    1.504369e-01      0
#> 66   -1.599571e+00      0
#> 67    1.495169e+00      0
#> 68    1.733330e-01      0
#> 69   -8.030292e-01      0
#> 70    1.893532e-01      0
#> 71   -7.470132e-01      0
#> 72    6.387651e-01      0
#> 73   -1.688960e+00      0
#> 74    1.128321e+00      0
#> 75    2.447905e-01      0
#> 76   -1.723234e-01      0
#> 77   -2.321095e+00      0
#> 78   -5.548692e-01      0
#> 79    2.080995e+00      0
#> 80    2.757957e-01      0
#> 81   -1.785246e-01      0
#> 82    1.762243e+00      0
#> 83   -4.245992e-01      0
#> 84    1.033467e+00      0
#> 85    4.698312e-02      0
#> 86    4.070547e-01      0
#> 87   -4.768928e-01      0
#> 88    2.424519e-01      0
#> 89    5.707448e-01      0
#> 90    2.397438e+00      0
#> 91   -1.224808e+00      0
#> 92   -5.140057e-01      0
#> 93    2.222914e+00      0
#> 94    6.478926e-01      0
#> 95   -1.267688e+00      0
#> 96   -2.908471e-01      0
#> 97    1.587607e+00      0
#> 98    1.943877e-01      0
#> 99    1.604102e+00      0
#> 100   2.227075e-02      0
#> 101  -6.206953e-01      0
#> 102  -4.614687e-01      0
#> 103  -1.721601e+00      0
#> 104   1.427806e+00      0
#> 105  -1.721705e+00      0
#> 106  -1.377397e-01      0
#> 107   9.038840e-01      0
#> 108   9.629864e-01      0
#> 109  -2.226607e+00      0
#> 110  -1.110504e+00      0
#> 111  -7.252003e-01      0
#> 112  -1.212017e+00      0
#> 113   5.094389e-01      0
#> 114  -9.169414e-01      0
#> 115   4.252182e-01      0
#> 116   8.891026e-01      0
#> 117  -3.128396e-01      0
#> 118   1.140951e+00      0
#> 119  -1.999772e+00      0
#> 120  -5.536533e-01      0
#> 121   9.709630e-01      0
#> 122   1.217379e+00      0
#> 123   1.808327e-01      0
#> 124   5.225504e-01      0
#> 125  -7.739478e-01      0
#> 126   5.872050e-01      0
#> 127  -1.665522e+00      0
#> 128  -1.327824e-01      0
#> 129   1.518836e-01      0
#> 130  -1.093852e+00      0
#> 131  -2.703294e-01      0
#> 132  -2.815902e-01      0
#> 133  -1.586519e+00      0
#> 134  -5.297530e-01      0
#> 135  -1.506378e+00      0
#> 136  -7.978446e-01      0
#> 137   7.015410e-01      0
#> 138   1.510359e+00      0
#> 139   3.258789e-01      0
#> 140  -1.419235e+00      0
#> 141   1.748516e+00      0
#> 142  -1.295136e+00      0
#> 143  -2.352314e-01      0
#> 144  -1.777296e-01      0
#> 145  -9.370736e-01      0
#> 146   3.863313e-01      0
#> 147  -2.573729e+00      0
#> 148  -1.163045e+00      0
#> 149   1.518499e+00      0
#> 150   5.921046e-01      0
#> 151   7.486628e-02      0
#> 152   3.769282e-01      0
#> 153   1.063075e+00      0
#> 154  -1.687170e+00      0
#> 155   9.475670e-02      0
#> 156   1.902115e+00      0
#> 157  -7.069555e-01      0
#> 158   3.033444e-01      0
#> 159  -3.275120e-01      0
#> 160   9.934966e-01      0
#> 161  -6.370789e-02      0
#> 162   1.195258e-01      0
#> 163   7.636106e-02      0
#> 164  -1.955828e+00      0
#> 165   2.189981e-02      0
#> 166   1.673840e-01      0
#> 167   5.559963e-01      0
#> 168  -1.909379e+00      0
#> 169   4.696710e-01      0
#> 170  -5.162614e-01      0
#> 171   1.915959e+00      0
#> 172  -1.135001e+00      0
#> 173  -1.663948e+00      0
#> 174   5.636022e-01      0
#> 175  -7.523931e-01      0
#> 176  -3.955679e-01      0
#> 177   2.508891e-02      0
#> 178   1.585683e+00      0
#> 179  -8.192122e-01      0
#> 180   8.424916e-01      0
#> 181  -1.039330e+00      0
#> 182  -3.007665e-01      0
#> 183   5.473579e-01      0
#> 184  -1.052361e+00      0
#> 185  -1.908050e+00      0
#> 186   9.905804e-02      0
#> 187   8.685394e-01      0
#> 188   1.378813e+00      0
#> 189  -1.818902e+00      0
#> 190   1.356250e+00      0
#> 191   5.881543e-01      0
#> 192  -3.088127e-01      0
#> 193   9.939397e-02      0
#> 194   2.270051e+00      0
#> 195   1.500143e+00      0
#> 196  -1.210303e+00      0
#> 197  -1.963135e-01      0
#> 198  -2.773128e-01      0
#> 199   3.607471e-01      0
#> 200   6.543552e-01      0
#> 201   7.590109e-01      0
#> 202   1.958043e+00      0
#> 203   2.990831e-01      0
#> 204  -1.027364e+00      0
#> 205   2.801629e+00      0
#> 206  -3.680869e-01      0
#> 207   8.119674e-01      0
#> 208   1.004188e+00      0
#> 209   1.174022e+00      0
#> 210  -1.012610e+00      0
#> 211  -5.636556e-01      0
#> 212  -6.611297e-01      0
#> 213   8.526376e-01      0
#> 214   1.051110e+00      0
#> 215  -9.606532e-01      0
#> 216  -3.084974e-01      0
#> 217  -6.379805e-01      0
#> 218   3.679805e-01      0
#> 219  -1.912208e+00      0
#> 220   3.764633e-01      0
#> 221  -2.382746e-02      0
#> 222  -1.159802e+00      0
#> 223  -1.510468e+00      0
#> 224  -2.206683e+00      0
#> 225  -4.039711e-01      0
#> 226  -3.559872e-01      0
#> 227  -3.073895e-01      0
#> 228   7.916580e-01      0
#> 229  -9.857713e-01      0
#> 230   3.941563e-01      0
#> 231  -9.336122e-01      0
#> 232  -5.018024e-01      0
#> 233  -8.834976e-01      0
#> 234   1.370118e+00      0
#> 235   1.449864e-02      0
#> 236  -8.903881e-01      0
#> 237   2.881474e-01      0
#> 238  -3.593481e-02      0
#> 239   1.553667e-01      0
#> 240   1.934949e+00      0
#> 241   1.955301e+00      0
#> 242   2.939160e-01      0
#> 243   4.063681e-01      0
#> 244   1.378525e+00      0
#> 245  -5.269358e-01      0
#> 246   2.028572e+00      0
#> 247  -1.668131e+00      0
#> 248   2.246990e+00      0
#> 249  -1.217388e+00      0
#> 250   1.030243e+00      0
#> 251   5.820193e-01      0
#> 252  -6.539287e-01      0
#> 253   1.447627e+00      0
#> 254   2.939994e-01      0
#> 255   8.838280e-01      0
#> 256   1.017907e+00      0
#> 257  -8.718126e-02      0
#> 258   9.478686e-01      0
#> 259   3.961794e-01      0
#> 260   1.766518e-01      0
#> 261  -7.264547e-01      0
#> 262  -4.739210e-01      0
#> 263   1.293285e+00      0
#> 264   8.013402e-01      0
#> 265  -3.642431e-01      0
#> 266  -1.428923e+00      0
#> 267  -3.349986e-01      0
#> 268  -8.503444e-01      0
#> 269   5.126287e-01      0
#> 270  -6.375626e-02      0
#> 271   2.138091e+00      0
#> 272   1.145477e+00      0
#> 273  -2.703171e+00      0
#> 274  -7.318959e-01      0
#> 275  -2.093116e-01      0
#> 276   1.782880e+00      0
#> 277   9.103000e-01      0
#> 278  -9.385636e-01      0
#> 279   1.143215e+00      0
#> 280  -2.043025e+00      0
#> 281  -2.677948e-01      0
#> 282   5.528497e-02      0
#> 283   1.565427e+00      0
#> 284   1.513537e+00      0
#> 285   5.886020e-01      0
#> 286   1.797668e-01      0
#> 287  -4.850422e-01      0
#> 288  -4.912899e-01      0
#> 289  -1.048439e+00      0
#> 290   1.253780e+00      0
#> 291  -5.231956e-01      0
#> 292   1.417738e+00      0
#> 293  -6.060501e-01      0
#> 294  -9.907064e-01      0
#> 295  -2.787964e-01      0
#> 296  -1.392411e+00      0
#> 297  -1.473241e+00      0
#> 298  -1.829898e-02      0
#> 299   9.646826e-01      0
#> 300   3.978325e-01      0
#> 301  -1.039376e+00      0
#> 302   9.181624e-01      0
#> 303  -2.848986e-01      0
#> 304  -3.393561e-01      0
#> 305  -6.167391e-01      0
#> 306  -1.030203e-01      0
#> 307  -3.177633e-01      0
#> 308  -6.985637e-01      0
#> 309  -3.642259e-01      0
#> 310   1.100168e-01      0
#> 311  -1.972995e+00      0
#> 312   4.876768e-01      0
#> 313  -1.043336e+00      0
#> 314  -5.642817e-01      0
#> 315  -1.202716e-01      0
#> 316   7.947784e-01      0
#> 317   1.112751e+00      0
#> 318   5.336967e-01      0
#> 319   7.906284e-01      0
#> 320   9.950216e-02      0
#> 321  -1.001033e+00      0
#> 322   2.612546e-01      0
#> 323  -3.036301e-01      0
#> 324   6.487154e-01      0
#> 325  -7.595233e-01      0
#> 326   1.769559e+00      0
#> 327  -5.470736e-01      0
#> 328  -2.662772e-03      0
#> 329   3.830064e-01      0
#> 330  -2.517948e-01      0
#> 331  -7.254351e-02      0
#> 332   2.077448e+00      0
#> 333   7.950226e-01      0
#> 334  -2.601810e-01      0
#> 335   1.410356e-01      0
#> 336   5.483723e-01      0
#> 337   1.870029e-01      0
#> 338   1.870362e+00      0
#> 339   7.625477e-01      0
#> 340   5.850986e-01      0
#> 341   5.980349e-01      0
#> 342  -9.826225e-01      0
#> 343  -1.426589e+00      0
#> 344   1.948852e-01      0
#> 345  -1.948911e-01      0
#> 346  -6.824200e-01      0
#> 347  -4.409632e-01      0
#> 348   3.274731e-01      0
#> 349   1.898370e-01      0
#> 350   7.551143e-01      0
#> 351   2.819618e-01      0
#> 352   3.143418e-01      0
#> 353   5.471975e-01      0
#> 354  -2.622991e-01      0
#> 355   1.303605e+00      0
#> 356  -6.278465e-01      0
#> 357   9.282590e-01      0
#> 358  -4.726536e-01      0
#> 359  -1.197157e+00      0
#> 360  -4.724025e-01      0
#> 361  -5.747421e-02      0
#> 362   1.099352e+00      0
#> 363  -2.756746e-01      0
#> 364  -1.206302e+00      0
#> 365  -5.817855e-02      0
#> 366   1.119440e+00      0
#> 367  -1.129393e+00      0
#> 368  -1.953627e+00      0
#> 369  -4.030727e-01      0
#> 370  -2.183843e-01      0
#> 371  -2.384686e+00      0
#> 372   2.257374e+00      0
#> 373  -1.882556e+00      0
#> 374  -1.458537e+00      0
#> 375   4.157956e-01      0
#> 376   1.057678e+00      0
#> 377   7.445250e-01      0
#> 378   1.858631e+00      0
#> 379   2.609419e-01      0
#> 380   1.637552e-01      0
#> 381  -1.118084e+00      0
#> 382   1.954875e-01      0
#> 383   9.139476e-01      0
#> 384   9.279808e-02      0
#> 385  -1.566346e+00      0
#> 386  -6.183996e-01      0
#> 387   1.077554e+00      0
#> 388   4.686492e-01      0
#> 389   3.111982e-01      0
#> 390   6.242284e-01      0
#> 391   6.816027e-01      0
#> 392   1.478720e+00      0
#> 393   5.287772e-01      0
#> 394  -4.553886e-01      0
#> 395   5.439180e-01      0
#> 396  -3.556187e-01      0
#> 397  -2.170228e+00      0
#> 398   1.282351e+00      0
#> 399   6.587396e-01      0
#> 400   2.913529e-01      0
#> 401  -3.100553e-01      0
#> 402   7.604982e-02      0
#> 403  -1.649524e+00      0
#> 404   6.445578e-01      0
#> 405   1.899808e-01      0
#> 406  -1.487527e+00      0
#> 407  -7.561680e-01      0
#> 408   1.147501e+00      0
#> 409  -1.651719e+00      0
#> 410   2.518925e+00      0
#> 411   7.176202e-01      0
#> 412   3.031442e-01      0
#> 413  -4.059539e-01      0
#> 414  -1.725007e+00      0
#> 415   1.212596e+00      0
#> 416   1.817181e+00      0
#> 417  -3.305142e-05      0
#> 418   5.216065e-01      0
#> 419   1.903534e+00      0
#> 420  -8.490564e-01      0
#> 421   6.818243e-01      0
#> 422   5.987655e-02      0
#> 423  -9.487831e-01      0
#> 424   1.061802e+00      0
#> 425   8.664899e-01      0
#> 426  -6.138612e-01      0
#> 427   7.518307e-01      0
#> 428  -9.502400e-01      0
#> 429   1.163200e+00      0
#> 430  -9.044713e-01      0
#> 431   1.250024e+00      0
#> 432  -6.895934e-01      0
#> 433   1.356264e+00      0
#> 434  -1.454130e-01      0
#> 435  -2.270443e-01      0
#> 436   6.033162e-01      0
#> 437   4.283886e-02      0
#> 438  -1.244630e+00      0
#> 439  -2.715132e+00      0
#> 440  -2.715057e+00      0
#> 441   7.138157e-01      0
#> 442   1.369156e+00      0
#> 443   9.615173e-01      0
#> 444  -7.985987e-02      0
#> 445   9.571024e-01      0
#> 446   1.545273e+00      0
#> 447   1.776309e+00      0
#> 448   8.950599e-01      0
#> 449   7.619251e-01      0
#> 450  -1.211850e+00      0
#> 451  -2.454648e+00      0
#> 452   7.868557e-01      0
#> 453   5.830428e-01      0
#> 454   4.226956e-01      0
#> 455  -1.834345e+00      0
#> 456   2.050113e-01      0
#> 457  -6.445699e-01      0
#> 458   1.108029e-02      0
#> 459   5.261937e-01      0
#> 460   1.910068e+00      0
#> 461  -3.043117e-01      0
#> 462  -8.869525e-01      0
#> 463  -3.025510e-01      0
#> 464  -1.405061e+00      0
#> 465   6.374208e-01      0
#> 466   8.906187e-01      0
#> 467   1.084392e+00      0
#> 468  -5.803842e-02      0
#> 469  -3.881443e-01      0
#> 470  -8.733449e-04      0
#> 471   4.157965e-01      0
#> 472  -1.127007e-01      0
#> 473  -1.862484e+00      0
#> 474  -1.598235e+00      0
#> 475  -1.380922e+00      0
#> 476   6.831699e-01      0
#> 477  -1.199800e-01      0
#> 478   1.715635e-01      0
#> 479  -5.971075e-01      0
#> 480  -2.562342e-01      0
#> 481   1.399413e+00      0
#> 482  -5.158910e-02      0
#> 483  -8.610582e-01      0
#> 484  -6.836565e-01      0
#> 485  -1.673523e+00      0
#> 486  -1.205893e-01      0
#> 487   3.278225e-01      0
#> 488  -7.721356e-01      0
#> 489   4.930946e-01      0
#> 490   4.076966e-01      0
#> 491  -2.707258e-01      0
#> 492   5.525033e-01      0
#> 493  -1.509429e-01      0
#> 494  -1.538014e-02      0
#> 495   6.589676e-01      0
#> 496  -6.564122e-01      0
#> 497  -8.012896e-01      0
#> 498   1.178160e+00      0
#> 499  -8.049405e-01      0
#> 500   7.644605e-01      0
#> 501   1.137975e+00      0
#> 502  -1.617836e+00      0
#> 503   5.217948e-01      0
#> 504  -2.462310e-01      0
#> 505   7.957094e-01      0
#> 506  -1.593117e+00      0
#> 507  -2.125152e-01      0
#> 508   1.549249e+00      0
#> 509  -1.662091e+00      0
#> 510   4.194139e-02      0
#> 511   1.583495e-01      0
#> 512  -3.236225e-02      0
#> 513  -7.681654e-01      0
#> 514  -1.070563e+00      0
#> 515   4.222613e-01      0
#> 516  -3.520131e-01      0
#> 517  -1.056998e+00      0
#> 518   1.740847e+00      0
#> 519  -5.758560e-01      0
#> 520  -4.535992e-01      0
#> 521   1.015268e+00      0
#> 522  -3.850136e-01      0
#> 523  -1.920448e-02      0
#> 524  -2.067326e+00      0
#> 525   2.642249e-01      0
#> 526  -1.506482e-01      0
#> 527   2.203167e+00      0
#> 528   2.078918e+00      0
#> 529  -2.132898e-01      0
#> 530   5.811306e-01      0
#> 531  -7.537607e-02      0
#> 532  -4.915499e-01      0
#> 533   1.723006e+00      0
#> 534   2.550307e-01      0
#> 535  -7.103934e-02      0
#> 536  -9.613452e-01      0
#> 537   5.158422e-01      0
#> 538  -1.086463e+00      0
#> 539   6.898698e-01      0
#> 540   9.531529e-01      0
#> 541   1.613817e+00      0
#> 542   2.179154e-01      0
#> 543   7.123393e-01      0
#> 544   4.654169e-01      0
#> 545  -2.647125e-01      0
#> 546  -1.579206e+00      0
#> 547  -9.907254e-01      0
#> 548  -9.763091e-01      0
#> 549   1.216136e+00      0
#> 550  -2.916688e-01      0
#> 551   8.091550e-01      0
#> 552   2.934851e-01      0
#> 553  -3.724292e-01      0
#> 554  -3.435906e-01      0
#> 555  -1.921879e+00      0
#> 556  -1.097748e+00      0
#> 557  -1.419313e+00      0
#> 558   3.673593e-01      0
#> 559   9.836157e-01      0
#> 560   3.736500e-01      0
#> 561  -2.039253e+00      0
#> 562  -6.797701e-01      0
#> 563  -4.835062e-01      0
#> 564  -7.334354e-02      0
#> 565  -9.880995e-01      0
#> 566   1.227773e+00      0
#> 567  -7.956381e-01      0
#> 568   3.567388e-01      0
#> 569   3.625107e-01      0
#> 570   2.229442e+00      0
#> 571  -5.181198e-01      0
#> 572  -9.164065e-01      0
#> 573   9.779976e-01      0
#> 574  -6.652693e-01      0
#> 575   4.991095e-01      0
#> 576  -2.284523e+00      0
#> 577  -5.527034e-01      0
#> 578  -3.247282e-01      0
#> 579  -6.411220e-01      0
#> 580   7.595596e-01      0
#> 581  -1.049117e+00      0
#> 582  -1.220920e+00      0
#> 583   3.582961e-01      0
#> 584  -8.033421e-01      0
#> 585   6.126847e-01      0
#> 586   8.342629e-01      0
#> 587  -1.388224e-01      0
#> 588   1.746296e+00      0
#> 589  -1.089700e-01      0
#> 590   1.377131e-01      0
#> 591  -8.700803e-02      0
#> 592   1.422759e+00      0
#> 593   1.220418e+00      0
#> 594  -2.557717e-02      0
#> 595   1.735925e+00      0
#> 596  -2.801373e-01      0
#> 597   1.547374e+00      0
#> 598   1.869814e+00      0
#> 599   1.122028e+00      0
#> 600   1.069006e-01      0
#> 601   5.822256e-01      0
#> 602  -7.367302e-01      0
#> 603  -2.928705e-01      0
#> 604   7.502332e-01      0
#> 605  -8.526574e-01      0
#> 606   1.775545e+00      0
#> 607   1.127046e+00      0
#> 608  -3.150469e-01      0
#> 609  -3.596764e-01      0
#> 610   3.551291e-01      0
#> 611   8.096390e-01      0
#> 612   1.766172e+00      0
#> 613   4.371187e-01      0
#> 614  -1.308775e+00      0
#> 615  -1.537019e-01      0
#> 616  -3.148984e-01      0
#> 617   1.912598e+00      0
#> 618  -1.400994e+00      0
#> 619   5.043334e-01      0
#> 620   8.801968e-01      0
#> 621   7.712012e-01      0
#> 622   7.139859e-03      0
#> 623   1.002815e+00      0
#> 624  -9.645488e-01      0
#> 625  -2.368087e-02      0
#> 626  -2.088929e+00      0
#> 627  -1.434036e-01      0
#> 628  -7.749282e-01      0
#> 629   7.014586e-01      0
#> 630  -7.141329e-01      0
#> 631   8.915169e-01      0
#> 632   7.372848e-02      0
#> 633  -4.974741e-01      0
#> 634   4.970030e-01      0
#> 635  -1.089047e+00      0
#> 636  -3.610505e-01      0
#> 637   1.048538e+00      0
#> 638  -6.824789e-01      0
#> 639   1.614416e-01      0
#> 640  -1.322702e-01      0
#> 641  -6.105734e-01      0
#> 642   1.206234e+00      0
#> 643   5.342349e-01      0
#> 644  -2.412545e-01      0
#> 645   1.163098e+00      0
#> 646  -1.589519e+00      0
#> 647   2.322241e-01      0
#> 648  -3.294478e-01      0
#> 649  -1.082619e+00      0
#> 650   6.974369e-02      0
#> 651  -2.875913e-01      0
#> 652   9.583721e-01      0
#> 653   1.801610e+00      0
#> 654  -1.205279e+00      0
#> 655   3.731051e-02      0
#> 656   9.811582e-01      0
#> 657   1.405804e-01      0
#> 658  -1.334902e+00      0
#> 659  -2.414655e-01      0
#> 660  -4.663628e-01      0
#> 661  -1.290347e+00      0
#> 662  -1.141985e+00      0
#> 663   2.724966e-01      0
#> 664   7.149650e-01      0
#> 665  -4.015148e-01      0
#> 666  -1.406284e+00      0
#> 667   8.482645e-01      0
#> 668  -3.413946e-02      0
#> 669   4.633979e-02      0
#> 670   1.342609e+00      0
#> 671  -1.740153e+00      0
#> 672  -4.411447e-01      0
#> 673  -6.692900e-01      0
#> 674  -8.568203e-01      0
#> 675   1.029604e+00      0
#> 676  -2.868173e-01      0
#> 677   1.096757e+00      0
#> 678  -7.978165e-01      0
#> 679   1.576968e+00      0
#> 680   6.553974e-01      0
#> 681   2.520081e+00      0
#> 682   6.753030e-01      0
#> 683  -1.613186e+00      0
#> 684   1.550877e+00      0
#> 685  -1.097269e+00      0
#> 686   9.584810e-01      0
#> 687   1.272063e-01      0
#> 688   4.075207e-01      0
#> 689  -1.214848e+00      0
#> 690  -3.700395e-02      0
#> 691  -1.570062e+00      0
#> 692  -5.163339e-01      0
#> 693  -3.128463e-01      0
#> 694  -5.157401e-01      0
#> 695  -6.489977e-01      0
#> 696   8.702038e-01      0
#> 697  -7.291987e-02      0
#> 698  -1.220934e+00      0
#> 699  -1.481945e+00      0
#> 700  -1.776705e+00      0
#> 701  -3.657396e-01      0
#> 702   3.036855e-01      0
#> 703  -1.379125e+00      0
#> 704  -6.592563e-02      0
#> 705   7.658605e-01      0
#> 706   3.546734e-01      0
#> 707   9.220837e-01      0
#> 708  -1.136975e+00      0
#> 709  -9.464097e-01      0
#> 710   1.052354e+00      0
#> 711  -7.436023e-02      0
#> 712  -1.368298e+00      0
#> 713  -1.919434e+00      0
#> 714  -2.621675e-02      0
#> 715  -9.036320e-01      0
#> 716   4.994891e-01      0
#> 717   1.194686e+00      0
#> 718   8.672392e-01      0
#> 719   1.077529e+00      0
#> 720   2.343316e-01      0
#> 721  -1.847675e+00      0
#> 722  -5.049304e-01      0
#> 723  -5.204339e-01      0
#> 724   1.555029e+00      0
#> 725  -2.452891e-01      0
#> 726  -3.756243e-01      0
#> 727   1.393614e+00      0
#> 728  -9.885424e-02      0
#> 729   5.223032e-01      0
#> 730   6.355993e-01      0
#> 731   1.809886e+00      0
#> 732   1.472239e+00      0
#> 733  -1.375302e-01      0
#> 734  -8.271885e-01      0
#> 735   1.421921e+00      0
#> 736   1.115896e+00      0
#> 737  -5.103202e-02      0
#> 738  -3.092473e-01      0
#> 739   3.678180e-02      0
#> 740  -7.537828e-01      0
#> 741  -4.364913e-01      0
#> 742   3.247704e-01      0
#> 743   2.917385e+00      0
#> 744  -5.930155e-01      0
#> 745  -7.189076e-02      0
#> 746   1.155477e-01      0
#> 747  -5.083735e-01      0
#> 748   1.463976e+00      0
#> 749  -2.100002e-01      0
#> 750   1.758519e-01      0
#> 751   7.490452e-01      0
#> 752   2.363505e+00      0
#> 753  -1.045623e+00      0
#> 754  -1.449313e+00      0
#> 755  -3.948967e-01      0
#> 756   1.093766e+00      0
#> 757   8.005189e-01      0
#> 758  -1.012797e+00      0
#> 759  -1.670946e+00      0
#> 760  -5.252677e-01      0
#> 761  -1.976660e-01      0
#> 762   1.881023e+00      0
#> 763  -4.979708e-01      0
#> 764   5.303820e-01      0
#> 765  -9.659621e-01      0
#> 766   3.266664e-01      0
#> 767   1.880561e-01      0
#> 768  -7.107798e-01      0
#> 769  -5.003730e-01      0
#> 770   1.409996e+00      0
#> 771  -4.288366e-02      0
#> 772  -9.482841e-01      0
#> 773  -9.286505e-01      0
#> 774  -1.192270e+00      0
#> 775   1.771434e+00      0
#> 776  -3.034329e-02      0
#> 777   1.974596e+00      0
#> 778  -8.416446e-02      0
#> 779   6.393219e-01      0
#> 780  -4.495211e-02      0
#> 781  -2.442968e-01      0
#> 782  -1.317852e+00      0
#> 783  -9.696455e-01      0
#> 784  -1.897436e+00      0
#> 785   7.595536e-01      0
#> 786   7.765148e-01      0
#> 787  -1.734741e+00      0
#> 788   5.683070e-01      0
#> 789  -3.801770e-01      0
#> 790   1.443988e+00      0
#> 791  -1.095874e+00      0
#> 792   2.766680e+00      0
#> 793   4.891276e-01      0
#> 794  -1.806764e-02      0
#> 795   4.525296e-01      0
#> 796  -2.921879e-01      0
#> 797   1.744134e+00      0
#> 798   7.563261e-01      0
#> 799   2.219612e-01      0
#> 800   3.554368e-01      0
#> 801  -6.768562e-01      0
#> 802  -4.864782e-01      0
#> 803  -1.807461e+00      0
#> 804   6.996178e-01      0
#> 805   1.115596e-01      0
#> 806  -5.975846e-01      0
#> 807   1.113122e+00      0
#> 808   1.377527e+00      0
#> 809   5.249312e-01      0
#> 810  -6.684204e-01      0
#> 811  -5.596258e-01      0
#> 812  -4.760819e-01      0
#> 813   1.362871e+00      0
#> 814   1.529427e+00      0
#> 815  -1.193236e+00      0
#> 816  -1.074886e+00      0
#> 817  -1.227750e+00      0
#> 818  -2.753803e-01      0
#> 819  -1.853055e+00      0
#> 820   6.849533e-01      0
#> 821  -2.022318e+00      0
#> 822   1.973917e+00      0
#> 823  -2.189959e-01      0
#> 824   8.391283e-01      0
#> 825   6.346711e-01      0
#> 826  -7.228944e-02      0
#> 827   2.152127e+00      0
#> 828  -5.591081e-01      0
#> 829   1.095899e+00      0
#> 830   2.606251e-01      0
#> 831   3.852314e-01      0
#> 832   1.804667e-01      0
#> 833   1.012772e+00      0
#> 834  -1.206150e-01      0
#> 835  -4.430055e-01      0
#> 836  -2.540863e-01      0
#> 837  -1.597104e+00      0
#> 838  -5.443256e-01      0
#> 839  -2.810611e-01      0
#> 840   1.215676e+00      0
#> 841   5.144789e-02      0
#> 842  -1.022963e+00      0
#> 843  -2.186896e-01      0
#> 844   4.659763e-01      0
#> 845  -4.554409e-02      0
#> 846  -6.125612e-01      0
#> 847   1.690516e+00      0
#> 848   1.014821e+00      0
#> 849  -3.145914e-02      0
#> 850  -6.672524e-01      0
#> 851   6.037922e-01      0
#> 852   1.203590e+00      0
#> 853   2.505924e-01      0
#> 854   1.594527e+00      0
#> 855   6.692710e-01      0
#> 856   1.156551e-01      0
#> 857   4.601169e-01      0
#> 858   7.458848e-01      0
#> 859  -1.365121e-01      0
#> 860  -8.654016e-01      0
#> 861  -4.800045e-01      0
#> 862   1.224562e+00      0
#> 863  -1.529456e+00      0
#> 864  -3.226578e-03      0
#> 865  -1.846897e-01      0
#> 866   5.341981e-01      0
#> 867   5.119616e-01      0
#> 868   9.265208e-01      0
#> 869   2.090762e+00      0
#> 870   1.266096e-01      0
#> 871   1.923544e-01      0
#> 872   5.847717e-01      0
#> 873   7.699992e-01      0
#> 874   3.484726e+00      0
#> 875  -5.493262e-01      0
#> 876   3.472603e-01      0
#> 877   1.459197e+00      0
#> 878   1.575251e+00      0
#> 879   2.024628e-01      0
#> 880   6.289227e-01      0
#> 881  -6.510904e-01      0
#> 882   2.144082e+00      0
#> 883   2.491563e-01      0
#> 884  -1.946968e-01      0
#> 885  -3.445368e-02      0
#> 886   1.289935e+00      0
#> 887  -1.839605e-01      0
#> 888  -3.144141e-01      0
#> 889  -6.412857e-01      0
#> 890   7.261859e-01      0
#> 891   5.086896e-01      0
#> 892   2.527074e+00      0
#> 893  -9.012082e-01      0
#> 894  -4.299180e-01      0
#> 895   9.912608e-01      0
#> 896  -2.972364e-01      0
#> 897  -6.747663e-01      0
#> 898   1.632487e+00      0
#> 899  -3.642778e-01      0
#> 900   4.048985e-01      0
#> 901  -2.649071e+00      0
#> 902  -8.800882e-02      0
#> 903  -1.305990e+00      0
#> 904   5.067041e-01      0
#> 905   2.811564e+00      0
#> 906   2.478459e-01      0
#> 907  -3.027641e-01      0
#> 908   7.063532e-01      0
#> 909   6.894420e-01      0
#> 910  -1.156669e+00      0
#> 911  -9.046495e-02      0
#> 912  -8.104394e-01      0
#> 913  -2.100527e+00      0
#> 914  -1.410807e+00      0
#> 915   4.385295e-02      0
#> 916   6.421944e-01      0
#> 917   8.603472e-01      0
#> 918  -1.534721e-01      0
#> 919   1.517578e+00      1
#> 920   1.272694e+00      1
#> 921   3.106402e-01      1
#> 922  -2.093813e-01      1
#> 923   8.663484e-01      1
#> 924   1.841833e+00      1
#> 925  -5.806164e-01      1
#> 926   1.845162e+00      1
#> 927   3.174282e-01      1
#> 928   1.430623e-01      1
#> 929  -9.731743e-03      1
#> 930   4.880980e-03      1
#> 931   1.156502e+00      1
#> 932   1.626767e+00      1
#> 933   2.696384e-01      1
#> 934   6.948755e-01      1
#> 935   3.247953e+00      1
#> 936   8.187796e-01      1
#> 937   2.086384e+00      1
#> 938   2.975542e+00      1
#> 939  -8.590867e-02      1
#> 940   2.140884e+00      1
#> 941   6.313537e-01      1
#> 942   2.976426e+00      1
#> 943  -5.961390e-01      1
#> 944   2.127085e+00      1
#> 945   2.709378e-02      1
#> 946   9.126092e-01      1
#> 947   3.349378e-01      1
#> 948   1.431393e+00      1
#> 949   5.095150e-02      1
#> 950   2.323077e+00      1
#> 951   6.656610e-01      1
#> 952  -1.239675e-01      1
#> 953   1.027499e+00      1
#> 954  -2.420095e+00      1
#> 955   8.528636e-01      1
#> 956  -2.910997e-01      1
#> 957   3.899959e-01      1
#> 958   1.172700e+00      1
#> 959  -1.105142e+00      1
#> 960   3.075616e-01      1
#> 961   4.040412e-01      1
#> 962  -5.769966e-01      1
#> 963   2.990765e+00      1
#> 964   1.501222e+00      1
#> 965   1.795346e-01      1
#> 966   8.869014e-01      1
#> 967   1.218448e+00      1
#> 968   6.761358e-01      1
#> 969   6.856678e-01      1
#> 970   1.154017e+00      1
#> 971   7.862568e-01      1
#> 972   3.457094e-01      1
#> 973   1.334799e+00      1
#> 974   3.386867e+00      1
#> 975  -7.262312e-01      1
#> 976  -3.860325e-01      1
#> 977   1.662454e-01      1
#> 978   3.578440e-01      1
#> 979   5.919194e-01      1
#> 980   1.905625e+00      1
#> 981  -5.474003e-01      1
#> 982   1.426702e+00      1
#> 983  -7.010076e-01      1
#> 984  -2.428805e-01      1
#> 985   5.330177e-01      1
#> 986   1.007298e-01      1
#> 987   1.622536e+00      1
#> 988   6.613512e-01      1
#> 989   2.324599e-01      1
#> 990  -7.740734e-02      1
#> 991   1.444852e+00      1
#> 992   1.401837e+00      1
#> 993   2.197146e+00      1
#> 994   9.455494e-01      1
#> 995   1.157101e+00      1
#> 996   3.677803e-01      1
#> 997   1.263670e+00      1
#> 998   1.199918e+00      1
#> 999  -7.752622e-01      1
#> 1000 -1.450743e+00      1