Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     0.3740298517      0
#> 2    -1.1151803248      0
#> 3     0.6860606220      0
#> 4    -0.4293782902      0
#> 5     0.1719988058      0
#> 6    -0.5321837167      0
#> 7    -1.6761706681      0
#> 8    -0.1158900002      0
#> 9     0.5429095948      0
#> 10   -0.1933293894      0
#> 11   -0.1628675091      0
#> 12   -0.1870573185      0
#> 13   -0.2044779209      0
#> 14   -1.1313189341      0
#> 15    0.9559401382      0
#> 16    0.0035288180      0
#> 17   -0.5458034502      0
#> 18    1.8237440982      0
#> 19    0.7612550345      0
#> 20   -1.6259556188      0
#> 21    2.4100825893      0
#> 22    0.0539105361      0
#> 23    0.1638724855      0
#> 24    0.0935925114      0
#> 25   -0.8399833166      0
#> 26    0.3014846061      0
#> 27   -0.7548563055      0
#> 28    0.3687843048      0
#> 29   -0.2283867367      0
#> 30   -1.9134606334      0
#> 31    0.8966615895      0
#> 32    1.1126219071      0
#> 33    1.0623416458      0
#> 34   -0.9615918401      0
#> 35   -0.1177249059      0
#> 36   -1.6978813204      0
#> 37    0.1883013603      0
#> 38    0.5251610218      0
#> 39   -0.6057181140      0
#> 40   -0.2319868153      0
#> 41   -0.7814781288      0
#> 42    0.4423114394      0
#> 43    2.2312811626      0
#> 44    1.4011446767      0
#> 45   -0.2910552335      0
#> 46   -0.9803291683      0
#> 47    0.5286759894      0
#> 48    0.6130361266      0
#> 49   -0.4518221503      0
#> 50   -0.0434116851      0
#> 51    1.7826800750      0
#> 52    0.1676201272      0
#> 53   -0.2008288809      0
#> 54   -0.7608583396      0
#> 55    0.0402528106      0
#> 56   -0.7559993629      0
#> 57   -0.1619198751      0
#> 58    2.4834176478      0
#> 59    1.0816980204      0
#> 60    1.6048048955      0
#> 61    0.1607148687      0
#> 62    1.9576754609      0
#> 63   -0.3080666656      0
#> 64    0.4970137425      0
#> 65    1.9373433324      0
#> 66    1.0479331599      0
#> 67    3.9555850375      0
#> 68   -0.2794213626      0
#> 69   -0.0119148421      0
#> 70   -0.5267570984      0
#> 71    1.4617135013      0
#> 72    0.2798538935      0
#> 73    0.4315075413      0
#> 74   -2.3571892243      0
#> 75    2.3304148475      0
#> 76    0.2337011446      0
#> 77    0.9800295470      0
#> 78    1.0614426571      0
#> 79    0.0075695831      0
#> 80   -0.3644855772      0
#> 81    0.0307589684      0
#> 82   -0.2979772853      0
#> 83    0.3074962165      0
#> 84   -1.0546597825      0
#> 85   -1.5949732813      0
#> 86    0.3712690407      0
#> 87    0.7315396067      0
#> 88   -0.9709313767      0
#> 89   -1.7755863291      0
#> 90    0.5640965086      0
#> 91   -0.7261909387      0
#> 92    0.8512794144      0
#> 93    0.0807893280      0
#> 94    0.9485159450      0
#> 95   -0.7206471491      0
#> 96    0.2010037703      0
#> 97    0.0562265577      0
#> 98    0.7091917637      0
#> 99   -0.0714277490      0
#> 100  -0.1350510443      0
#> 101  -0.0160391645      0
#> 102   0.4763755532      0
#> 103   0.0652299558      0
#> 104  -1.3528706277      0
#> 105   1.3927013674      0
#> 106   1.2248959060      0
#> 107  -0.2951994241      0
#> 108   0.1856404031      0
#> 109   0.4713020586      0
#> 110   1.3800369343      0
#> 111   0.5692313075      0
#> 112   1.0745614727      0
#> 113   0.6758847210      0
#> 114   0.2450698716      0
#> 115  -0.7969524906      0
#> 116   0.5154669702      0
#> 117   0.4603586346      0
#> 118   1.0770553481      0
#> 119   0.7143581133      0
#> 120  -0.5075661557      0
#> 121  -0.5423438919      0
#> 122   0.1696909148      0
#> 123  -0.3593555575      0
#> 124   1.2222954226      0
#> 125  -1.1757684234      0
#> 126  -0.6706322507      0
#> 127   0.2950709128      0
#> 128  -1.1231536041      0
#> 129  -0.2403333940      0
#> 130   0.6814130253      0
#> 131   1.3420649080      0
#> 132  -1.2291013612      0
#> 133   0.6459719520      0
#> 134   1.1853862223      0
#> 135   0.0073992395      0
#> 136  -0.9034140931      0
#> 137   0.7733485065      0
#> 138  -1.1888482366      0
#> 139  -0.5476942124      0
#> 140  -0.0886480392      0
#> 141  -0.2547979862      0
#> 142  -0.3728328681      0
#> 143   0.0867731929      0
#> 144  -0.8103833407      0
#> 145   0.2891527092      0
#> 146   2.0812652170      0
#> 147  -0.7710354775      0
#> 148  -0.2038696650      0
#> 149  -0.9422489724      0
#> 150  -1.0151897257      0
#> 151  -2.9332576729      0
#> 152   0.6519088613      0
#> 153  -1.0612251995      0
#> 154  -0.0471968458      0
#> 155  -0.1957426801      0
#> 156   1.7119531134      0
#> 157  -0.0337077084      0
#> 158  -0.6191693260      0
#> 159  -0.5109096787      0
#> 160  -2.6632041269      0
#> 161  -0.6322140994      0
#> 162   0.6377673293      0
#> 163  -0.7645131079      0
#> 164   0.6215170781      0
#> 165  -0.4841208261      0
#> 166  -0.8001770347      0
#> 167   0.5727701375      0
#> 168   0.6077158957      0
#> 169  -0.6315281470      0
#> 170  -0.2569057734      0
#> 171   1.1170363875      0
#> 172  -0.2097160212      0
#> 173   0.3087423890      0
#> 174  -0.5204522885      0
#> 175  -0.2094343184      0
#> 176  -0.4069205577      0
#> 177   0.3521494856      0
#> 178  -0.8021924640      0
#> 179  -1.6517999749      0
#> 180  -0.6209029045      0
#> 181   0.5876692398      0
#> 182   0.3851640928      0
#> 183  -0.0709778164      0
#> 184  -1.8360527017      0
#> 185   0.6307277256      0
#> 186  -1.0157289193      0
#> 187  -0.5914570992      0
#> 188   0.0564239594      0
#> 189  -0.4771883963      0
#> 190  -0.1367496930      0
#> 191  -0.7322832212      0
#> 192   1.1725910882      0
#> 193  -0.3224755362      0
#> 194   0.4736523357      0
#> 195   0.3086992352      0
#> 196   0.9680542340      0
#> 197   1.1322238147      0
#> 198  -1.4934740563      0
#> 199  -0.5228622468      0
#> 200  -0.3164771536      0
#> 201   0.4685235254      0
#> 202  -1.1146959416      0
#> 203   0.1189582032      0
#> 204  -0.8495102576      0
#> 205  -0.5813929459      0
#> 206  -0.9666559458      0
#> 207  -1.2897123911      0
#> 208   1.5045147622      0
#> 209  -0.1462981905      0
#> 210  -0.4977675395      0
#> 211   0.4817061678      0
#> 212   1.2228956312      0
#> 213   2.5905145571      0
#> 214  -2.3547614648      0
#> 215   0.5802420078      0
#> 216   0.9590997677      0
#> 217   0.0253972302      0
#> 218  -0.0024468517      0
#> 219  -0.8596348676      0
#> 220   1.2740456287      0
#> 221  -1.7064146907      0
#> 222   1.0937493669      0
#> 223  -0.0950654260      0
#> 224  -1.2213575089      0
#> 225  -1.2741990433      0
#> 226  -1.5671405128      0
#> 227  -0.7771443862      0
#> 228   0.1571005531      0
#> 229   0.2705831477      0
#> 230   1.1147698701      0
#> 231   0.6257599155      0
#> 232   1.6806683348      0
#> 233  -0.5948159824      0
#> 234  -0.7752718490      0
#> 235  -2.3921735535      0
#> 236   0.5920435218      0
#> 237  -0.1083465776      0
#> 238  -0.2686739194      0
#> 239  -0.1513712748      0
#> 240  -1.3514096933      0
#> 241   0.1809086954      0
#> 242  -0.9411060079      0
#> 243   0.1261315255      0
#> 244  -0.1880195544      0
#> 245  -0.5203108990      0
#> 246  -0.7136638439      0
#> 247  -0.1968475004      0
#> 248   0.8396321720      0
#> 249  -0.8640027244      0
#> 250  -1.8921561304      0
#> 251   0.2404385485      0
#> 252   1.2844681817      0
#> 253  -0.1298728448      0
#> 254  -0.3347775329      0
#> 255  -1.7115146596      0
#> 256  -0.1670132272      0
#> 257   0.0792572315      0
#> 258  -1.2657630220      0
#> 259  -0.7037462879      0
#> 260   1.2973483090      0
#> 261   1.1076813605      0
#> 262   0.5066955161      0
#> 263  -0.0773018685      0
#> 264  -2.0760450709      0
#> 265  -0.6407752455      0
#> 266   0.2994934566      0
#> 267  -0.1024860203      0
#> 268   1.4183728339      0
#> 269   0.1166473785      0
#> 270  -0.9149280518      0
#> 271  -0.8999557636      0
#> 272   0.5537457370      0
#> 273  -1.2900145346      0
#> 274  -0.8996558089      0
#> 275   1.8024787161      0
#> 276  -0.6768684788      0
#> 277   1.0214290081      0
#> 278  -0.2970234055      0
#> 279  -1.1405297098      0
#> 280  -0.3975653849      0
#> 281  -0.4907374115      0
#> 282  -0.0205341552      0
#> 283   2.8832324962      0
#> 284  -0.8993910146      0
#> 285   0.3034706955      0
#> 286  -0.9305458655      0
#> 287   0.4456969312      0
#> 288  -0.7335327750      0
#> 289  -0.0938316480      0
#> 290  -0.1917125621      0
#> 291   0.2522213821      0
#> 292  -1.7032724514      0
#> 293   1.5012409250      0
#> 294  -0.1541252130      0
#> 295  -0.1778547246      0
#> 296   0.4655579343      0
#> 297   0.3380405159      0
#> 298  -1.2754758670      0
#> 299  -0.3323312599      0
#> 300   1.2466893972      0
#> 301  -1.4568061168      0
#> 302  -0.9391839421      0
#> 303   1.1452102179      0
#> 304  -0.6836674549      0
#> 305  -0.8952765598      0
#> 306   0.9532202642      0
#> 307   0.4277607609      0
#> 308  -0.0618248715      0
#> 309   0.5756311325      0
#> 310  -0.4911153010      0
#> 311   0.2684985409      0
#> 312  -0.3896697682      0
#> 313   0.8836007582      0
#> 314  -0.2849347871      0
#> 315  -0.3167390968      0
#> 316  -0.5002483639      0
#> 317  -0.4945966187      0
#> 318  -0.1643786270      0
#> 319  -2.1259234365      0
#> 320   0.8279827521      0
#> 321  -0.3436681760      0
#> 322   1.6788591710      0
#> 323   0.5553960470      0
#> 324   0.6767521955      0
#> 325  -1.3479997266      0
#> 326   0.8074541887      0
#> 327  -0.2189112955      0
#> 328   0.5771302419      0
#> 329   1.0342315829      0
#> 330  -1.3767183811      0
#> 331   0.1477550467      0
#> 332   0.3707989679      0
#> 333   0.0810245262      0
#> 334  -0.7512345276      0
#> 335  -0.1305893826      0
#> 336  -0.2941517309      0
#> 337   0.2984293921      0
#> 338   0.1240454775      0
#> 339  -1.0306571269      0
#> 340  -0.4847328569      0
#> 341   0.0257374964      0
#> 342  -1.1816465234      0
#> 343  -1.2383698337      0
#> 344   0.0131575126      0
#> 345   0.5027425158      0
#> 346   1.7379188719      0
#> 347   1.5220298639      0
#> 348   1.8355072013      0
#> 349  -0.2782186926      0
#> 350  -1.4969493452      0
#> 351  -0.7048626243      0
#> 352   0.1289980808      0
#> 353  -0.9079265483      0
#> 354   0.3661267655      0
#> 355  -0.1402101624      0
#> 356   1.3432708528      0
#> 357   1.1568049341      0
#> 358   0.5191427274      0
#> 359  -2.6858775866      0
#> 360   0.4463164757      0
#> 361   1.1432961490      0
#> 362  -0.0321016434      0
#> 363  -1.3910803763      0
#> 364  -0.8127211020      0
#> 365  -0.5138926909      0
#> 366   0.9911285646      0
#> 367  -0.9371928496      0
#> 368  -1.8721543209      0
#> 369   0.2095602557      0
#> 370  -0.2344768787      0
#> 371   0.7936037283      0
#> 372  -0.7454165139      0
#> 373  -0.1357048096      0
#> 374   0.3803950871      0
#> 375   0.4223334607      0
#> 376   0.4462094609      0
#> 377  -0.3486805117      0
#> 378   1.3995980462      0
#> 379  -0.0514781377      0
#> 380   0.0426789937      0
#> 381   0.1686603550      0
#> 382  -1.9178846386      0
#> 383  -1.7329349321      0
#> 384  -1.2967697223      0
#> 385  -0.4564598473      0
#> 386  -0.6386390558      0
#> 387  -0.9103571103      0
#> 388  -0.0911898440      0
#> 389   1.1009921919      0
#> 390  -0.5842013116      0
#> 391   1.9729157085      0
#> 392  -0.5818054307      0
#> 393  -0.1638115433      0
#> 394   1.3305608559      0
#> 395   1.4883220480      0
#> 396  -0.3436497812      0
#> 397  -0.5610447155      0
#> 398  -0.0619584273      0
#> 399   0.4184385421      0
#> 400   1.6543934157      0
#> 401   1.7346030514      0
#> 402   1.3122790758      0
#> 403   0.4833764353      0
#> 404   0.2672642908      0
#> 405  -0.3463195325      0
#> 406   0.4068273733      0
#> 407  -0.5091582767      0
#> 408   0.8236886140      0
#> 409   0.5114978980      0
#> 410   0.6950474098      0
#> 411  -0.6677229177      0
#> 412  -0.6539278607      0
#> 413  -0.9684474233      0
#> 414  -1.7039951717      0
#> 415  -0.5213403391      0
#> 416   1.9609488524      0
#> 417  -0.7792290128      0
#> 418  -1.1171808646      0
#> 419  -0.6616050842      0
#> 420   0.0982188581      0
#> 421  -0.3483811589      0
#> 422   0.3955043536      0
#> 423  -0.2230422590      0
#> 424   0.0568516587      0
#> 425  -0.1371592000      0
#> 426   1.5174240572      0
#> 427   0.3224215188      0
#> 428   0.3199091801      0
#> 429   0.5002069003      0
#> 430   1.3833634303      0
#> 431   0.2421312652      0
#> 432   0.0166992507      0
#> 433  -0.6539289499      0
#> 434  -0.1251355693      0
#> 435  -1.6927885778      0
#> 436  -0.9605322844      0
#> 437  -0.1239697797      0
#> 438   0.7382802090      0
#> 439  -1.1317479387      0
#> 440  -1.0487246456      0
#> 441   0.1062312789      0
#> 442   0.2321403495      0
#> 443   0.2249920615      0
#> 444  -0.8908116445      0
#> 445  -1.1031243074      0
#> 446  -0.8362413291      0
#> 447   0.3704344847      0
#> 448  -0.2378722812      0
#> 449  -1.0088468670      0
#> 450  -0.3107787861      0
#> 451   0.8938838448      0
#> 452   0.1080271516      0
#> 453   0.0388171695      0
#> 454   0.0297356155      0
#> 455   0.9328547518      0
#> 456  -0.0236277084      0
#> 457   0.9301767795      0
#> 458  -2.1413276059      0
#> 459   0.1507471161      0
#> 460   1.3814246174      0
#> 461  -0.2492521737      0
#> 462  -1.4993641345      0
#> 463   0.4696877427      0
#> 464   2.0433892352      0
#> 465   1.6554773204      0
#> 466  -0.9488664553      0
#> 467  -0.1555959104      0
#> 468  -0.6303708031      0
#> 469  -0.3118375548      0
#> 470  -0.1067222246      0
#> 471  -0.6044609790      0
#> 472   2.0857150145      0
#> 473   0.5148751563      0
#> 474  -0.5499783272      0
#> 475   1.1568437670      0
#> 476   0.0803151176      0
#> 477  -1.4981341587      0
#> 478   0.0609082194      0
#> 479   1.2923094229      0
#> 480  -0.4898635157      0
#> 481  -0.3541075793      0
#> 482  -0.1635562495      0
#> 483  -0.9177117599      0
#> 484   1.3029847312      0
#> 485   1.0820021738      0
#> 486   0.7774693279      0
#> 487  -0.1986775362      0
#> 488  -0.2238855268      0
#> 489  -0.1633060028      0
#> 490  -0.7920296540      0
#> 491  -0.4149339076      0
#> 492  -1.2170879646      0
#> 493  -0.3447783995      0
#> 494  -0.9681110877      0
#> 495  -0.6563601488      0
#> 496   0.3552074739      0
#> 497   1.2046992286      0
#> 498  -0.0677747277      0
#> 499  -0.6314808922      0
#> 500   0.4780294621      0
#> 501   0.9809335069      0
#> 502  -0.5259168741      0
#> 503   0.7408158684      0
#> 504  -1.0755161973      0
#> 505  -0.0842698315      0
#> 506  -1.1880629329      0
#> 507   1.0910311259      0
#> 508  -0.2605498195      0
#> 509   0.9082190765      0
#> 510   0.8345209246      0
#> 511  -0.1527390409      0
#> 512  -1.4468986244      0
#> 513  -0.1246920756      0
#> 514  -1.7090138102      0
#> 515  -0.7665697266      0
#> 516  -0.5469154794      0
#> 517   0.1160660382      0
#> 518  -0.1979083784      0
#> 519  -0.9588025459      0
#> 520   0.0908844160      0
#> 521  -0.5619410324      0
#> 522  -0.4201505942      0
#> 523  -0.0453132340      0
#> 524  -1.8870308880      0
#> 525   1.1201453594      0
#> 526   0.0609768851      0
#> 527  -1.7809201157      0
#> 528   0.1517632263      0
#> 529  -0.9501855067      0
#> 530   0.1273517481      0
#> 531  -0.7976266689      0
#> 532  -0.6977534730      0
#> 533  -0.6540041862      0
#> 534   0.7771571142      0
#> 535   1.7703835060      0
#> 536  -2.2612734587      0
#> 537  -0.7622322512      0
#> 538  -0.6865021382      0
#> 539  -0.7934981030      0
#> 540  -0.3091593316      0
#> 541  -1.7191941501      0
#> 542  -0.3988300014      0
#> 543  -0.5224485079      0
#> 544   0.6884592495      0
#> 545   0.5635997663      0
#> 546   0.7879572961      0
#> 547   0.4696328710      0
#> 548   2.2753813156      0
#> 549  -0.0622414758      0
#> 550   2.1410956813      0
#> 551  -0.4214460727      0
#> 552   0.0265456505      0
#> 553  -0.5596127671      0
#> 554  -0.1985052599      0
#> 555  -0.5906147313      0
#> 556  -2.2627849548      0
#> 557  -0.3021729195      0
#> 558   1.3261938061      0
#> 559   0.4131492119      0
#> 560   0.7896105412      0
#> 561  -0.1322606857      0
#> 562  -2.4669659733      0
#> 563  -1.4905284211      0
#> 564   0.1357212204      0
#> 565   1.9940129015      0
#> 566   2.6302698988      0
#> 567   1.0769623390      0
#> 568  -2.7404966425      0
#> 569   0.4651716834      0
#> 570   0.8050749494      0
#> 571  -0.3619330605      0
#> 572  -0.0346632268      0
#> 573   0.6081386262      0
#> 574  -0.5893000825      0
#> 575  -1.1418111427      0
#> 576   1.2551702889      0
#> 577   1.4676872496      0
#> 578  -0.3021203284      0
#> 579  -0.6070795860      0
#> 580   1.3369394641      0
#> 581   1.7389903963      0
#> 582  -1.8136163841      0
#> 583   0.8721755522      0
#> 584   0.4485944628      0
#> 585  -1.4512567506      0
#> 586   0.3190395740      0
#> 587  -0.1246063966      0
#> 588   1.7261383200      0
#> 589   0.7370312694      0
#> 590  -1.4462844127      0
#> 591   0.2626760905      0
#> 592   0.6988152573      0
#> 593   1.0082116993      0
#> 594   0.5982531087      0
#> 595  -0.2413625635      0
#> 596   1.6039204401      0
#> 597   0.2183501042      0
#> 598  -1.1560964732      0
#> 599   1.0295970549      0
#> 600   1.2596293613      0
#> 601  -1.4450073611      0
#> 602   0.9161749360      0
#> 603  -0.4662548971      0
#> 604   1.1451946547      0
#> 605   0.1350077130      0
#> 606  -0.5845173021      0
#> 607  -0.8327710612      0
#> 608  -0.0295548469      0
#> 609  -0.2536520939      0
#> 610   1.4432595001      0
#> 611  -0.3690298596      0
#> 612   0.8311966475      0
#> 613   0.8720239718      0
#> 614   1.4109853212      0
#> 615   0.2510770579      0
#> 616  -0.3202089578      0
#> 617   0.0299347228      0
#> 618  -0.9410596363      0
#> 619   0.4154368097      0
#> 620  -1.2139865913      0
#> 621   3.4041814032      0
#> 622  -0.8532256609      0
#> 623   0.0741196615      0
#> 624   0.4865325959      0
#> 625  -0.4165936524      0
#> 626  -0.3174835004      0
#> 627  -0.1146196067      0
#> 628  -1.2495537023      0
#> 629   0.8720471292      0
#> 630  -0.0280949617      0
#> 631  -1.6230121219      0
#> 632  -0.1474509670      0
#> 633  -0.5682722180      0
#> 634  -0.6544417313      0
#> 635   0.1548126397      0
#> 636  -1.0823027033      0
#> 637  -0.1293456716      0
#> 638   0.5621345282      0
#> 639  -1.0275569116      0
#> 640  -0.7671197788      0
#> 641   0.1365230293      0
#> 642  -0.3473274655      0
#> 643  -1.7816863814      0
#> 644   0.5008907350      0
#> 645  -0.4947598623      0
#> 646  -0.5379144160      0
#> 647   0.9676356894      0
#> 648   0.1409990303      0
#> 649   0.1416168416      0
#> 650  -0.3318712281      0
#> 651  -1.2483120451      0
#> 652  -1.0475711621      0
#> 653   1.3103573144      0
#> 654  -0.0704530935      0
#> 655   0.4810399667      0
#> 656  -0.3988985569      0
#> 657   0.2711733089      0
#> 658  -1.2062602510      0
#> 659   0.0057804909      0
#> 660   0.7735989826      0
#> 661  -1.2297414072      0
#> 662  -2.2327849458      0
#> 663   0.0189824240      0
#> 664  -1.1652640193      0
#> 665   0.5921510526      0
#> 666   0.4625249692      0
#> 667   0.2830254259      0
#> 668   1.1227288884      0
#> 669  -0.5755042857      0
#> 670  -1.2662390839      0
#> 671  -1.0590982091      0
#> 672  -0.5765534094      0
#> 673   0.6978067336      0
#> 674   0.0831268088      0
#> 675  -1.1603978213      0
#> 676  -3.1070257354      0
#> 677  -1.3301282149      0
#> 678   1.4750200091      0
#> 679  -1.3941079275      0
#> 680   0.0298878893      0
#> 681  -0.3038233150      0
#> 682   0.2230046285      0
#> 683   0.9179082905      0
#> 684  -0.5294082041      0
#> 685  -1.5109204950      0
#> 686  -0.8260173027      0
#> 687  -1.4090498872      0
#> 688   1.4464103765      0
#> 689   0.3363570633      0
#> 690  -0.4988230115      0
#> 691  -0.5267422598      0
#> 692   0.8011767203      0
#> 693   0.2313482097      0
#> 694   0.0500084657      0
#> 695   0.5651999097      0
#> 696   0.2624883736      0
#> 697   2.8284731195      0
#> 698  -0.4044074586      0
#> 699  -0.7146470474      0
#> 700   0.0180983321      0
#> 701   0.1728955362      0
#> 702  -0.8186310225      0
#> 703  -1.6204762947      0
#> 704  -0.3776972935      0
#> 705   0.7042572929      0
#> 706   0.7140347050      0
#> 707   1.6288809289      0
#> 708   0.7312587137      0
#> 709   2.1464284216      0
#> 710  -0.1180545051      0
#> 711   0.0397439594      0
#> 712  -0.7497722095      0
#> 713   0.0158205946      0
#> 714  -0.9271733293      0
#> 715   1.7914509674      0
#> 716   0.1967607841      0
#> 717  -1.1233613849      0
#> 718   1.1007650607      0
#> 719   0.6405680134      0
#> 720  -0.6616298593      0
#> 721  -1.6242810540      0
#> 722  -0.4504430003      0
#> 723  -0.9460607090      0
#> 724   0.6201802619      0
#> 725   0.5837293002      0
#> 726  -0.2362079216      0
#> 727   1.3000033113      0
#> 728   1.4945235864      0
#> 729   1.6433905328      0
#> 730   0.5006836502      0
#> 731  -0.4591545503      0
#> 732   0.0900280906      0
#> 733  -2.0504330347      0
#> 734   0.5964805450      0
#> 735  -0.8340625377      0
#> 736  -0.6553520666      0
#> 737  -0.5445115705      0
#> 738  -0.2686959172      0
#> 739   0.4877036700      0
#> 740   0.1919622192      0
#> 741  -0.8031685530      0
#> 742   1.5753517430      0
#> 743  -0.0549792896      0
#> 744  -1.0445540976      0
#> 745  -0.4334927231      0
#> 746   1.2867876492      0
#> 747   0.9646486719      0
#> 748   1.1794884146      0
#> 749  -0.0557503449      0
#> 750  -0.9825223699      0
#> 751  -1.1774510944      0
#> 752  -1.1773365479      0
#> 753  -1.4249002338      0
#> 754   2.2358596366      0
#> 755  -0.1137071342      0
#> 756  -0.1461489363      0
#> 757  -2.9218678990      0
#> 758  -0.4072085342      0
#> 759   1.0930693737      0
#> 760  -2.4120838885      0
#> 761  -0.3587182239      0
#> 762   0.6740640494      0
#> 763  -0.3950457507      0
#> 764   0.6278214026      0
#> 765  -1.1785250352      0
#> 766  -1.7378315744      0
#> 767  -0.0612139696      0
#> 768  -0.5125699686      0
#> 769  -0.0375476033      0
#> 770   3.3005229913      0
#> 771  -0.7698033841      0
#> 772  -0.0573712208      0
#> 773  -0.0121212725      0
#> 774  -0.2102940448      0
#> 775   1.2800145000      0
#> 776   1.9867521960      0
#> 777   0.4469403129      0
#> 778  -0.5848713733      0
#> 779   0.6378522687      0
#> 780  -1.9703950675      0
#> 781   0.5378051532      0
#> 782  -0.9749315718      0
#> 783   1.4444365585      0
#> 784   0.5970613687      0
#> 785  -0.0558123645      0
#> 786   0.6741049921      0
#> 787  -1.5576289780      0
#> 788   0.0127813627      0
#> 789   1.6433180963      0
#> 790   0.0995074172      0
#> 791   0.9751024977      0
#> 792   1.5062316810      0
#> 793  -0.7266517117      0
#> 794   1.1510017106      0
#> 795   0.2855806862      0
#> 796  -1.7572666494      0
#> 797   0.2575453119      0
#> 798   0.6784300132      0
#> 799  -0.3213983347      0
#> 800   0.7592705597      0
#> 801  -0.2113775129      0
#> 802   0.7005492616      0
#> 803   0.4416529602      0
#> 804  -0.9727827899      0
#> 805  -0.6140498378      0
#> 806  -1.0028157031      0
#> 807   0.1190704887      0
#> 808  -0.6921956177      0
#> 809  -0.3187533335      0
#> 810   1.1149825836      0
#> 811  -0.8301908313      0
#> 812  -1.0901102617      0
#> 813  -1.1565273705      0
#> 814  -0.7112730089      0
#> 815  -0.5221756563      0
#> 816   0.1319416916      0
#> 817   0.8611434755      0
#> 818  -0.2726135601      0
#> 819  -0.1179375451      0
#> 820  -0.6049798553      0
#> 821   0.4996913846      0
#> 822   0.9484046067      0
#> 823   1.2095719123      0
#> 824  -0.4859568452      0
#> 825   0.3939044540      0
#> 826  -0.7188519786      0
#> 827   1.0181835402      0
#> 828  -1.6979162094      0
#> 829  -0.4813541508      0
#> 830   1.2541091346      0
#> 831   0.4523576713      0
#> 832   0.5164289835      0
#> 833   0.1652212421      0
#> 834  -2.1906025497      0
#> 835  -0.5642823799      0
#> 836  -0.4855647526      0
#> 837  -0.0007947117      0
#> 838   0.4113184988      0
#> 839  -0.7761130370      0
#> 840  -0.1776586824      0
#> 841  -0.0734728520      0
#> 842   0.5724075611      0
#> 843   0.4066315659      0
#> 844  -0.3515463223      0
#> 845   0.2607935093      0
#> 846   0.9955643684      0
#> 847   2.2691803273      0
#> 848  -0.4879878974      0
#> 849   0.6757513966      0
#> 850   1.0832302994      0
#> 851   0.0802159616      0
#> 852  -0.4301348486      0
#> 853  -0.5200976868      0
#> 854  -0.3411567061      0
#> 855  -0.2354540771      0
#> 856  -0.3956434997      0
#> 857   0.8944797846      0
#> 858   0.5443777384      0
#> 859   1.4143524305      0
#> 860   0.6113621473      0
#> 861  -1.5070471959      0
#> 862  -0.1362850630      0
#> 863   0.7531572598      0
#> 864   0.0096510043      0
#> 865  -0.5188186993      0
#> 866   0.8524133690      0
#> 867   1.5034565067      0
#> 868  -1.4153020431      0
#> 869  -1.2884402389      0
#> 870   1.0628836550      0
#> 871  -0.2440204594      0
#> 872   1.5826194538      0
#> 873  -1.1063073759      0
#> 874   1.7261954648      0
#> 875   1.8222153869      0
#> 876  -0.8512122463      0
#> 877  -0.8897103608      0
#> 878   1.3059741400      0
#> 879   1.7747177558      0
#> 880   0.4986525463      0
#> 881   0.2045723499      0
#> 882  -1.7572386416      0
#> 883  -1.2472725946      0
#> 884  -1.0045568026      0
#> 885  -0.3698528470      0
#> 886   0.3355578196      0
#> 887   0.1064495079      0
#> 888  -0.0730724390      0
#> 889   0.5461708554      0
#> 890  -0.0140096922      0
#> 891   1.2161647363      0
#> 892   1.4186276192      1
#> 893   0.8034851102      1
#> 894  -0.1311764086      1
#> 895  -0.8435759513      1
#> 896   1.3281626946      1
#> 897   0.5223640026      1
#> 898   0.6322612126      1
#> 899   1.2026463406      1
#> 900   0.5459464732      1
#> 901   1.0764200104      1
#> 902  -1.0493242622      1
#> 903  -0.3327933635      1
#> 904   1.2683658881      1
#> 905  -0.5029224539      1
#> 906   1.6836126463      1
#> 907  -0.1998732385      1
#> 908   0.0858341401      1
#> 909   2.6097738962      1
#> 910   1.7604089036      1
#> 911   1.6966179104      1
#> 912   1.0467826483      1
#> 913   0.2183482238      1
#> 914  -0.3150416901      1
#> 915   0.6571315526      1
#> 916   0.9712152706      1
#> 917  -0.3067720987      1
#> 918  -0.2918426772      1
#> 919  -1.0942562546      1
#> 920  -0.4326992348      1
#> 921  -0.0389481521      1
#> 922   1.1157011322      1
#> 923   1.9188727778      1
#> 924   1.4307534545      1
#> 925  -0.1395007256      1
#> 926  -1.1192004155      1
#> 927   1.1421346301      1
#> 928   2.2770571381      1
#> 929  -0.8944174346      1
#> 930  -1.3699757581      1
#> 931   1.5681710084      1
#> 932  -0.3003044124      1
#> 933   1.3515899820      1
#> 934   1.6559504829      1
#> 935   0.0347821701      1
#> 936   1.3615788056      1
#> 937   0.6819928406      1
#> 938  -0.0641127634      1
#> 939   0.6947515761      1
#> 940   1.3336699375      1
#> 941   1.4545953663      1
#> 942   0.7283307555      1
#> 943   0.5217611340      1
#> 944   0.0134918365      1
#> 945  -0.3877065248      1
#> 946   0.5910155844      1
#> 947   1.3405087660      1
#> 948  -1.1479351594      1
#> 949  -1.6311332730      1
#> 950   0.2526430857      1
#> 951   0.2825803581      1
#> 952   0.7194828741      1
#> 953   1.0874406508      1
#> 954  -0.9543500103      1
#> 955   0.5019601087      1
#> 956  -2.0634177788      1
#> 957  -0.1332495747      1
#> 958   1.9404804034      1
#> 959   2.8937133272      1
#> 960   0.3140695146      1
#> 961  -0.1971122794      1
#> 962   0.5247969792      1
#> 963  -0.5675070837      1
#> 964   1.4874362818      1
#> 965  -1.0842404856      1
#> 966  -0.5391764045      1
#> 967   1.8058188474      1
#> 968   1.6680060912      1
#> 969   0.0595518129      1
#> 970   1.3048233304      1
#> 971   0.9767329800      1
#> 972   0.3073703858      1
#> 973   1.1873119593      1
#> 974  -0.9332448942      1
#> 975  -0.7686705945      1
#> 976   1.9746972422      1
#> 977   2.8646028322      1
#> 978   1.0608897813      1
#> 979   0.9190083196      1
#> 980   1.1860626369      1
#> 981   2.6300376217      1
#> 982   1.8607424168      1
#> 983   1.5241808372      1
#> 984   1.5039396986      1
#> 985   1.0699459466      1
#> 986   1.7617676966      1
#> 987  -2.7147404273      1
#> 988  -0.0263979779      1
#> 989   1.3193600616      1
#> 990   2.2929529770      1
#> 991   0.7957920962      1
#> 992   1.3811158752      1
#> 993   0.4988587993      1
#> 994  -0.0359268558      1
#> 995   0.8182272105      1
#> 996   0.2648576191      1
#> 997   1.5373872910      1
#> 998   2.6539162475      1
#> 999   0.2073415937      1
#> 1000  0.8889629804      1