Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     1.6917560806      0
#> 2    -0.8681588632      0
#> 3    -0.3690754182      0
#> 4    -1.1065133097      0
#> 5    -0.7129607836      0
#> 6    -0.6497944099      0
#> 7    -0.3781580126      0
#> 8     0.4590401037      0
#> 9     2.1491885698      0
#> 10    0.4822412826      0
#> 11   -1.6598364601      0
#> 12   -1.4153818940      0
#> 13   -1.9262424359      0
#> 14    1.8615245357      0
#> 15   -0.3754142488      0
#> 16   -1.7789349793      0
#> 17    0.6922041317      0
#> 18    0.6817146993      0
#> 19    1.2889299917      0
#> 20   -1.1340406558      0
#> 21   -0.5026894753      0
#> 22    1.4573077561      0
#> 23   -0.9037872459      0
#> 24    0.4709605390      0
#> 25    0.3987721173      0
#> 26   -0.0613539283      0
#> 27   -1.8742702627      0
#> 28    1.1033470900      0
#> 29   -0.5320909432      0
#> 30    0.1380457100      0
#> 31    1.1198671896      0
#> 32    1.0374966412      0
#> 33   -0.7943879015      0
#> 34   -0.1450848643      0
#> 35   -0.1850230513      0
#> 36   -1.7722613855      0
#> 37   -0.5456965468      0
#> 38   -0.8284071742      0
#> 39    0.3904815468      0
#> 40   -1.0201221630      0
#> 41    1.2522895167      0
#> 42    0.4824495738      0
#> 43    1.5310007362      0
#> 44   -0.9478707200      0
#> 45    1.1810709048      0
#> 46    1.5458325122      0
#> 47   -0.7696015025      0
#> 48   -1.2464921211      0
#> 49   -3.4175540094      0
#> 50    1.0674208429      0
#> 51   -0.4400950995      0
#> 52    1.2997549285      0
#> 53    0.9876461156      0
#> 54    0.8920811613      0
#> 55   -1.4662768546      0
#> 56    0.1764544284      0
#> 57    1.2995536041      0
#> 58    0.6055049055      0
#> 59    1.8975487546      0
#> 60    0.2887551998      0
#> 61    0.1826663920      0
#> 62    1.1152829301      0
#> 63   -1.1695144462      0
#> 64    1.3689773306      0
#> 65    0.3920329232      0
#> 66    0.7077883471      0
#> 67   -0.0074910052      0
#> 68    0.3284246500      0
#> 69    0.1129829398      0
#> 70   -1.3222901119      0
#> 71   -2.7399551449      0
#> 72   -0.7422120286      0
#> 73   -0.3093481335      0
#> 74    0.0827461174      0
#> 75   -0.1226119519      0
#> 76    1.8464766803      0
#> 77   -0.8847608487      0
#> 78   -0.5240793333      0
#> 79   -0.9960298368      0
#> 80    0.2477394209      0
#> 81   -0.4808701235      0
#> 82   -0.6291707203      0
#> 83   -1.3700559035      0
#> 84   -0.2335194575      0
#> 85   -0.3971040148      0
#> 86   -1.8145203554      0
#> 87   -0.2010941924      0
#> 88    0.6179079919      0
#> 89    0.0719304782      0
#> 90   -2.0779173853      0
#> 91    0.1135359594      0
#> 92   -1.8253470461      0
#> 93   -1.0208495145      0
#> 94    0.2911730550      0
#> 95    1.5104922879      0
#> 96   -1.0377546804      0
#> 97    0.5697751048      0
#> 98   -1.0466349282      0
#> 99    0.4330887849      0
#> 100  -1.7913661399      0
#> 101   0.9939838022      0
#> 102   0.7642571817      0
#> 103  -0.7049225753      0
#> 104  -0.4008727511      0
#> 105   1.6552087078      0
#> 106  -1.0595077511      0
#> 107  -0.2025703693      0
#> 108   1.0065776864      0
#> 109   0.7031523989      0
#> 110   1.3701348249      0
#> 111   1.7365921918      0
#> 112   1.9765970515      0
#> 113  -0.2558623083      0
#> 114  -1.1692540431      0
#> 115   1.1787790196      0
#> 116  -0.3445275028      0
#> 117   0.3269763605      0
#> 118   0.7744682319      0
#> 119   0.3634776048      0
#> 120  -0.5111338219      0
#> 121  -0.8894266699      0
#> 122  -0.6902164727      0
#> 123  -0.6335202402      0
#> 124   0.1517143987      0
#> 125   0.6953690847      0
#> 126  -0.2949570076      0
#> 127  -0.6678130793      0
#> 128  -0.6120097181      0
#> 129  -0.0733416066      0
#> 130   0.9552580339      0
#> 131   1.7544823991      0
#> 132  -2.0831056491      0
#> 133   1.1823121306      0
#> 134  -0.8028961632      0
#> 135   0.7236822302      0
#> 136  -1.3302861902      0
#> 137  -0.6991477220      0
#> 138   0.7406456237      0
#> 139  -0.5435212239      0
#> 140  -0.8151451195      0
#> 141   0.2123781077      0
#> 142   1.0137575210      0
#> 143  -1.7984393895      0
#> 144   0.6223558376      0
#> 145   1.1485305931      0
#> 146   0.9216537532      0
#> 147   2.1435567560      0
#> 148  -0.1907670242      0
#> 149   0.6891453848      0
#> 150   0.0154653170      0
#> 151  -0.3535520256      0
#> 152  -2.3745442399      0
#> 153  -0.7326377936      0
#> 154  -0.8826785902      0
#> 155  -1.0570469886      0
#> 156   0.8295301213      0
#> 157   1.5247124341      0
#> 158   0.6487069068      0
#> 159   1.1560499374      0
#> 160  -0.2387418453      0
#> 161  -0.8239605326      0
#> 162  -0.4656988181      0
#> 163   0.4519770346      0
#> 164  -1.1401838373      0
#> 165   0.9004176742      0
#> 166   0.6446026939      0
#> 167  -0.7814460234      0
#> 168  -1.2626263651      0
#> 169  -0.5086079237      0
#> 170  -0.5625968490      0
#> 171  -0.3128329762      0
#> 172  -1.3316286561      0
#> 173   0.0169334341      0
#> 174  -0.5777244716      0
#> 175  -0.7566568350      0
#> 176  -0.3265094369      0
#> 177  -0.4856335540      0
#> 178  -1.1456898857      0
#> 179   0.3032917701      0
#> 180   0.9032438911      0
#> 181   0.9581142165      0
#> 182   0.8287519083      0
#> 183  -0.9893515141      0
#> 184   0.3156971332      0
#> 185  -1.2634634407      0
#> 186   0.6081057743      0
#> 187  -0.9926397868      0
#> 188   0.7550458682      0
#> 189   0.5406555880      0
#> 190   0.9859282622      0
#> 191   0.1866909339      0
#> 192  -0.4021996516      0
#> 193  -0.7066639889      0
#> 194  -0.4747305275      0
#> 195   0.0332891511      0
#> 196  -0.1012818777      0
#> 197   0.1418039365      0
#> 198   0.0462521958      0
#> 199  -0.7041093180      0
#> 200   0.7993144290      0
#> 201   2.4363839272      0
#> 202   1.6072221752      0
#> 203   0.3449971470      0
#> 204  -0.3179522951      0
#> 205  -0.1420937379      0
#> 206  -1.3099547978      0
#> 207   0.9253297242      0
#> 208   0.9906634158      0
#> 209   0.4163481227      0
#> 210   0.2689286325      0
#> 211  -1.2842831669      0
#> 212  -1.0953121143      0
#> 213  -1.1402488296      0
#> 214   0.4183800124      0
#> 215   0.0534694949      0
#> 216  -0.0496005793      0
#> 217   0.4764357281      0
#> 218  -0.6881307552      0
#> 219   1.3920492790      0
#> 220   0.3557442556      0
#> 221   1.0362630509      0
#> 222  -0.4839116701      0
#> 223   0.7371910288      0
#> 224   0.3942626190      0
#> 225   0.4563191155      0
#> 226  -1.6413146882      0
#> 227   0.4048874352      0
#> 228   0.6875217859      0
#> 229   1.4615655730      0
#> 230   0.6011711896      0
#> 231   0.7106025411      0
#> 232   0.4908287472      0
#> 233  -1.6727606311      0
#> 234   0.3375559105      0
#> 235  -0.1253549418      0
#> 236   0.0973611007      0
#> 237  -0.8801354215      0
#> 238   0.8913794162      0
#> 239  -0.1955442003      0
#> 240  -1.6667435612      0
#> 241   0.4045471575      0
#> 242   1.3050045204      0
#> 243  -2.2712267461      0
#> 244   2.5949602219      0
#> 245  -1.7627500778      0
#> 246  -0.9293006723      0
#> 247   1.4093725371      0
#> 248   0.1686618182      0
#> 249  -0.1961439793      0
#> 250   1.2885606859      0
#> 251  -0.6794700684      0
#> 252   0.3597714261      0
#> 253   1.2226622672      0
#> 254  -1.2025513217      0
#> 255   0.8989413828      0
#> 256  -0.1137655064      0
#> 257   0.7919862152      0
#> 258   0.0308806030      0
#> 259  -0.9195196945      0
#> 260  -1.2858307220      0
#> 261  -0.6630123454      0
#> 262  -1.2830577403      0
#> 263   0.0921125430      0
#> 264  -0.6906870829      0
#> 265  -0.1816673591      0
#> 266   0.6071011563      0
#> 267   0.3422156393      0
#> 268  -1.5857458449      0
#> 269  -0.1776429668      0
#> 270  -0.7739476383      0
#> 271   0.3642726124      0
#> 272  -0.2558113231      0
#> 273  -0.6977706204      0
#> 274   1.1603571995      0
#> 275  -1.4534319589      0
#> 276  -1.8917872504      0
#> 277   1.1910527474      0
#> 278  -0.4491233863      0
#> 279  -0.1004312762      0
#> 280   1.8701074085      0
#> 281   1.5069904835      0
#> 282  -0.9767288533      0
#> 283  -0.0343778083      0
#> 284   1.8314399656      0
#> 285   0.5077223190      0
#> 286   0.3552243623      0
#> 287  -1.7118215988      0
#> 288   0.4829674242      0
#> 289  -0.0431074605      0
#> 290  -0.5067714948      0
#> 291   1.0807479148      0
#> 292   0.1900677763      0
#> 293   1.1058022497      0
#> 294  -1.4689461600      0
#> 295  -0.4464540701      0
#> 296   0.8774052209      0
#> 297   1.2559475256      0
#> 298  -2.0121980852      0
#> 299   0.4815600720      0
#> 300   0.6433507868      0
#> 301  -0.4975948127      0
#> 302  -0.2643082258      0
#> 303  -1.5098901745      0
#> 304   0.7936504008      0
#> 305  -0.8111249173      0
#> 306   0.1394848495      0
#> 307  -0.7548479771      0
#> 308   0.3193288717      0
#> 309   0.6015507997      0
#> 310   0.0106141375      0
#> 311   0.2697399471      0
#> 312   1.2595716602      0
#> 313   0.0274385850      0
#> 314  -0.3329101224      0
#> 315  -0.8700464099      0
#> 316  -2.4586196417      0
#> 317   0.2948774239      0
#> 318  -0.3341312484      0
#> 319  -0.6326087993      0
#> 320  -0.4558305736      0
#> 321   1.5611101694      0
#> 322   0.3598740826      0
#> 323   0.0342403390      0
#> 324  -0.2226543962      0
#> 325   1.0498495021      0
#> 326  -0.1536578781      0
#> 327  -0.6666314248      0
#> 328   0.0428430132      0
#> 329  -0.6719570638      0
#> 330   0.7207751110      0
#> 331  -0.1435012783      0
#> 332  -0.5413327548      0
#> 333  -0.4161710859      0
#> 334   2.7154422582      0
#> 335   0.5593462745      0
#> 336   1.0852452233      0
#> 337   0.0321762647      0
#> 338   1.2450526994      0
#> 339  -0.1132844009      0
#> 340   0.0098866506      0
#> 341  -0.0004424593      0
#> 342   1.0692975512      0
#> 343  -0.9117170449      0
#> 344   0.0226380195      0
#> 345  -0.0139125056      0
#> 346  -0.2588157740      0
#> 347  -0.7313869801      0
#> 348  -0.8974834886      0
#> 349  -0.2997918735      0
#> 350  -0.8605603369      0
#> 351  -2.7514562731      0
#> 352  -0.0485431030      0
#> 353  -0.4361876195      0
#> 354   1.0229153302      0
#> 355  -1.1877135666      0
#> 356  -0.2357761243      0
#> 357  -0.4708582902      0
#> 358  -0.2278470247      0
#> 359   0.4919632940      0
#> 360   0.8260641590      0
#> 361  -0.2941092219      0
#> 362  -1.2871852850      0
#> 363  -1.2691430438      0
#> 364  -0.7542451665      0
#> 365  -0.6633158020      0
#> 366  -0.1213486200      0
#> 367  -0.5826089823      0
#> 368   1.0330933446      0
#> 369  -0.1737348656      0
#> 370   0.1286153519      0
#> 371  -2.3175999243      0
#> 372   1.5056580594      0
#> 373  -0.3234906033      0
#> 374   1.2792227359      0
#> 375  -0.0050482121      0
#> 376  -0.0922553618      0
#> 377  -0.3798889066      0
#> 378  -1.6073820287      0
#> 379  -0.5423901668      0
#> 380  -0.8077393462      0
#> 381  -0.7828790785      0
#> 382  -0.0050727557      0
#> 383   0.9922745364      0
#> 384  -1.1620032248      0
#> 385   0.2611656623      0
#> 386   0.3400039928      0
#> 387  -0.8469949257      0
#> 388  -1.5188714910      0
#> 389   0.4157294548      0
#> 390   1.5806128591      0
#> 391   1.4000378515      0
#> 392   0.4294130704      0
#> 393   1.0665503267      0
#> 394  -0.3530359135      0
#> 395  -1.9168734251      0
#> 396   0.1253539998      0
#> 397  -0.2418491282      0
#> 398   0.4726151202      0
#> 399  -0.3660822129      0
#> 400   0.2075807949      0
#> 401  -0.3663664528      0
#> 402   0.2394737908      0
#> 403  -0.1944130425      0
#> 404  -0.4776162601      0
#> 405   1.6828696564      0
#> 406  -1.4641382735      0
#> 407  -1.3192436712      0
#> 408   0.9703768047      0
#> 409   0.7385087193      0
#> 410   0.5449469444      0
#> 411  -0.6054795219      0
#> 412   1.6916410067      0
#> 413   0.1625251376      0
#> 414   0.4940443397      0
#> 415  -0.0588941930      0
#> 416  -0.0731939686      0
#> 417  -0.9662431100      0
#> 418  -1.5022540623      0
#> 419   0.2690990269      0
#> 420  -1.7143775950      0
#> 421  -0.0708068788      0
#> 422  -2.4677210224      0
#> 423   2.6981797294      0
#> 424  -0.2948214889      0
#> 425   0.1438581244      0
#> 426  -0.8083534246      0
#> 427  -0.1392881387      0
#> 428   0.2668036081      0
#> 429  -1.5949618582      0
#> 430   0.3320746270      0
#> 431  -0.3009484682      0
#> 432   1.0534795081      0
#> 433   0.0280999788      0
#> 434  -0.0781242142      0
#> 435  -0.0653897870      0
#> 436   0.8900689683      0
#> 437   0.2097778108      0
#> 438   1.1923907332      0
#> 439   0.0714483149      0
#> 440   0.2059614611      0
#> 441  -0.3259882170      0
#> 442  -1.5753964994      0
#> 443  -0.2209167742      0
#> 444  -2.1293492845      0
#> 445  -2.0203813463      0
#> 446   0.4998233275      0
#> 447  -0.8294479952      0
#> 448  -0.6017748211      0
#> 449  -1.1130179133      0
#> 450   1.0975974176      0
#> 451   0.5864403879      0
#> 452   0.1898442884      0
#> 453  -1.0842474960      0
#> 454   0.4491932489      0
#> 455  -1.0130767325      0
#> 456   0.7959984468      0
#> 457   0.7688760270      0
#> 458  -1.1192010885      0
#> 459   0.5531080255      0
#> 460  -0.2151865295      0
#> 461   1.2034729183      0
#> 462   1.1771664423      0
#> 463   0.4161931063      0
#> 464   1.2625311614      0
#> 465  -0.6535681392      0
#> 466   0.7250194221      0
#> 467  -1.2522693274      0
#> 468   1.0504938301      0
#> 469  -1.2081730269      0
#> 470  -0.8483573899      0
#> 471  -1.1175610079      0
#> 472  -0.6515296911      0
#> 473   0.8798766999      0
#> 474  -1.1642356345      0
#> 475  -0.2776580422      0
#> 476  -0.3691431678      0
#> 477   1.2758548440      0
#> 478   0.9749906914      0
#> 479   0.6576231356      0
#> 480   2.1505639965      0
#> 481   0.4919471403      0
#> 482  -0.9712038082      0
#> 483  -1.5593251459      0
#> 484   0.6458614693      0
#> 485   0.1582276773      0
#> 486   1.4922754380      0
#> 487   1.7989731281      0
#> 488   1.6480130242      0
#> 489   0.2530220165      0
#> 490   0.5444830959      0
#> 491   0.1349709632      0
#> 492  -1.5679756339      0
#> 493  -0.0030317403      0
#> 494  -0.4085102127      0
#> 495  -0.1697376884      0
#> 496  -1.3029631295      0
#> 497   0.2765074665      0
#> 498  -1.1432628052      0
#> 499  -1.6912372624      0
#> 500  -1.0679448290      0
#> 501  -2.1801495718      0
#> 502   0.3238331502      0
#> 503  -0.9294351365      0
#> 504  -1.5893551028      0
#> 505  -0.0639473657      0
#> 506   0.3180187925      0
#> 507  -0.8330626048      0
#> 508   1.3218951835      0
#> 509   0.3131630197      0
#> 510  -0.5654374142      0
#> 511  -0.1120330420      0
#> 512   2.8111598326      0
#> 513  -1.2662980947      0
#> 514  -0.4584238698      0
#> 515  -0.3630173616      0
#> 516  -0.6063924214      0
#> 517   0.3472021763      0
#> 518  -0.2771806231      0
#> 519   0.0475526185      0
#> 520   1.8355802605      0
#> 521  -0.1042863072      0
#> 522  -0.5236889120      0
#> 523   1.4359670414      0
#> 524   0.0753580410      0
#> 525   1.6381033425      0
#> 526  -1.4084687933      0
#> 527   0.3297762403      0
#> 528  -0.8134851470      0
#> 529  -0.2188262839      0
#> 530  -0.1713163733      0
#> 531   0.8472067764      0
#> 532   0.8134295704      0
#> 533  -0.5756588612      0
#> 534  -1.3851551090      0
#> 535   0.3415763359      0
#> 536   0.0044736739      0
#> 537   0.9122534937      0
#> 538   1.4655368773      0
#> 539  -0.1586532667      0
#> 540   0.3093491899      0
#> 541  -0.4773491767      0
#> 542  -1.0092591910      0
#> 543   1.0085063326      0
#> 544   0.7288752401      0
#> 545  -0.1084388404      0
#> 546   0.4278959937      0
#> 547  -2.2322575351      0
#> 548   2.1316009685      0
#> 549  -0.1348129543      0
#> 550  -1.2132336066      0
#> 551  -1.5236932882      0
#> 552   1.1996210961      0
#> 553   0.7224786744      0
#> 554  -2.0967655845      0
#> 555  -0.1151824960      0
#> 556   0.7996921171      0
#> 557   0.5918633865      0
#> 558  -0.6707553616      0
#> 559  -0.3437273632      0
#> 560  -0.1011913254      0
#> 561  -0.2357803226      0
#> 562  -0.7045275750      0
#> 563  -1.9130545167      0
#> 564  -0.6774558331      0
#> 565  -1.2633318424      0
#> 566   1.0761850808      0
#> 567   0.1855599322      0
#> 568  -0.1569337985      0
#> 569  -1.4974952827      0
#> 570   0.6905927365      0
#> 571  -0.2242355949      0
#> 572   0.9980894949      0
#> 573  -0.0525390686      0
#> 574   0.5876034822      0
#> 575   0.0679510374      0
#> 576  -0.4111397252      0
#> 577   0.1276219472      0
#> 578  -1.2594884335      0
#> 579   0.0590506817      0
#> 580  -0.0058195150      0
#> 581   0.3123600967      0
#> 582  -0.8538915968      0
#> 583   0.2725715873      0
#> 584   0.3954674657      0
#> 585   1.4930494784      0
#> 586  -1.6814593404      0
#> 587  -0.0097085200      0
#> 588  -2.5227649138      0
#> 589   1.1431928464      0
#> 590   0.3007610361      0
#> 591   0.2425701906      0
#> 592  -0.4077054576      0
#> 593   0.2626390198      0
#> 594  -1.0298661216      0
#> 595  -3.2621359590      0
#> 596  -0.2052042462      0
#> 597   0.4372910409      0
#> 598  -1.1899412259      0
#> 599  -0.4594500489      0
#> 600  -1.2937575363      0
#> 601   0.2463858705      0
#> 602   0.9332637806      0
#> 603  -0.5849281934      0
#> 604   0.4823817225      0
#> 605  -1.6970838412      0
#> 606   0.2223394003      0
#> 607   0.8188190293      0
#> 608  -0.6471901725      0
#> 609  -0.7002436887      0
#> 610   0.9268440636      0
#> 611  -0.3694299908      0
#> 612   0.2412443367      0
#> 613   0.0388309813      0
#> 614   0.1199876589      0
#> 615   1.7637018036      0
#> 616  -0.7646673747      0
#> 617  -1.8545506603      0
#> 618   0.6138872322      0
#> 619  -0.8131572767      0
#> 620  -0.2012530713      0
#> 621  -0.7819329567      0
#> 622  -0.9937433903      0
#> 623   0.8650603691      0
#> 624   1.2006879848      0
#> 625  -0.3511444552      0
#> 626  -1.4213310963      0
#> 627   1.2717725565      0
#> 628   1.8307378702      0
#> 629   0.5223302427      0
#> 630  -0.4756902855      0
#> 631  -0.3441890562      0
#> 632   1.0569511919      0
#> 633  -0.3331016823      0
#> 634  -0.4575231802      0
#> 635   0.5663433284      0
#> 636  -0.2372127831      0
#> 637  -0.7096792017      0
#> 638  -1.0479119505      0
#> 639   0.1358671366      0
#> 640  -1.1353720536      0
#> 641   1.8872671422      0
#> 642  -1.6695232946      0
#> 643   0.0878632492      0
#> 644  -0.9514196849      0
#> 645  -1.1006442498      0
#> 646  -0.6298490036      0
#> 647   0.5516349993      0
#> 648  -1.2486059762      0
#> 649  -1.4197332042      0
#> 650  -0.9549597122      0
#> 651   1.1325205358      0
#> 652  -1.3213921621      0
#> 653   0.8659876093      0
#> 654  -0.6774876377      0
#> 655  -0.1915003398      0
#> 656  -0.0856319663      0
#> 657  -0.2307999919      0
#> 658  -1.4265713124      0
#> 659  -0.8834239646      0
#> 660   0.4457015388      0
#> 661   0.2239765160      0
#> 662  -0.0936089769      0
#> 663   0.4584314459      0
#> 664   0.6800738664      0
#> 665   1.5226372534      0
#> 666   0.3951410411      0
#> 667  -0.8467777239      0
#> 668  -0.8125928903      0
#> 669   1.2217817100      0
#> 670  -0.7149591588      0
#> 671   1.1469208652      0
#> 672  -0.0880661067      0
#> 673   0.5172939285      0
#> 674   0.1843581587      0
#> 675  -0.8550756882      0
#> 676  -1.3681591593      0
#> 677  -2.6011125002      0
#> 678  -0.3208159069      0
#> 679  -0.0127904427      0
#> 680  -0.3930782933      0
#> 681   1.1426203661      0
#> 682   0.7417897081      0
#> 683  -1.3584847413      0
#> 684   0.1274216772      0
#> 685   1.2303427693      0
#> 686   0.6459529154      0
#> 687  -0.8520452233      0
#> 688   1.2274981336      0
#> 689   0.0817611559      0
#> 690   0.8408874931      0
#> 691  -0.3454779843      0
#> 692   0.0782048686      0
#> 693  -0.8530205574      0
#> 694   1.9929920473      0
#> 695  -0.5138101059      0
#> 696  -0.2848244605      0
#> 697   1.5193901616      0
#> 698   0.0922723048      0
#> 699   1.3485725926      0
#> 700   0.1119951636      0
#> 701   1.0244651735      0
#> 702   1.6046489398      0
#> 703   1.6139684338      0
#> 704  -0.5424197869      0
#> 705  -0.0179490797      0
#> 706  -0.6890051028      0
#> 707   0.8107763488      0
#> 708  -0.7103548703      0
#> 709   0.8954172842      0
#> 710   1.3400319923      0
#> 711  -0.1184997523      0
#> 712  -1.1714702617      0
#> 713  -0.3569046376      0
#> 714  -0.0200606287      0
#> 715   0.7310126124      0
#> 716   0.9390878880      0
#> 717   1.7163118152      0
#> 718  -0.2816934672      0
#> 719  -1.1416187342      0
#> 720   0.5995573614      0
#> 721  -0.5822365533      0
#> 722   1.8214112554      0
#> 723  -1.1346434540      0
#> 724   0.0190599970      0
#> 725  -0.5769960116      0
#> 726  -0.0325942581      0
#> 727  -1.4338609538      0
#> 728   1.2570426682      0
#> 729   0.3150084495      0
#> 730  -0.7009696696      0
#> 731   0.4675270707      0
#> 732  -0.8988607786      0
#> 733   0.2077295685      0
#> 734  -2.1500147890      0
#> 735   2.3488888412      0
#> 736  -1.0313629576      0
#> 737   1.3737958382      0
#> 738  -0.4975634755      0
#> 739   0.6270619375      0
#> 740  -2.2303361591      0
#> 741   0.2406240021      0
#> 742  -1.6317683241      0
#> 743   0.1373292948      0
#> 744  -0.0131007567      0
#> 745  -0.1994174965      0
#> 746   0.8468610634      0
#> 747   1.0956394173      0
#> 748   0.3781478897      0
#> 749   0.1553965444      0
#> 750   0.4076884341      0
#> 751   1.3953121475      0
#> 752   1.2802191847      0
#> 753   0.4728299198      0
#> 754  -0.6934801489      0
#> 755  -0.9953800788      0
#> 756  -0.4850482711      0
#> 757   1.1588117194      0
#> 758   0.0515472287      0
#> 759  -0.3809191904      0
#> 760  -0.7521809400      0
#> 761   1.2533647219      0
#> 762  -1.2061039896      0
#> 763  -0.3133300584      0
#> 764  -0.2423958958      0
#> 765  -2.2626532699      0
#> 766  -0.4458582730      0
#> 767   0.4834761104      0
#> 768   0.6949736251      0
#> 769   0.6103110161      0
#> 770   1.7162473337      0
#> 771  -1.8764027820      0
#> 772  -0.3153171237      0
#> 773  -0.5729429785      0
#> 774   0.6485420273      0
#> 775   1.3885105136      0
#> 776  -0.1440947999      0
#> 777   0.2309495727      0
#> 778   0.6842914967      0
#> 779   1.9592906113      0
#> 780   2.0912994363      0
#> 781  -1.1058957085      0
#> 782  -0.7324151558      0
#> 783  -0.0385662459      0
#> 784  -1.1456237506      0
#> 785   2.7412716402      0
#> 786   1.6495067621      0
#> 787  -1.1588846011      0
#> 788  -0.9857595526      0
#> 789  -0.0659480276      0
#> 790  -1.4928135484      0
#> 791   0.4249502821      0
#> 792  -1.1462384899      0
#> 793  -0.5795560610      0
#> 794  -0.8423026230      0
#> 795  -0.9603551425      0
#> 796   0.1235789104      0
#> 797   0.5451789140      0
#> 798  -0.4499960894      0
#> 799  -0.2660409987      0
#> 800  -0.0218864583      0
#> 801   1.0167404829      0
#> 802   2.5139789087      0
#> 803  -0.4178723962      0
#> 804  -0.4716168976      0
#> 805  -1.4508678886      0
#> 806  -0.3115098854      0
#> 807   0.9443333659      0
#> 808   1.7311878520      0
#> 809  -0.6273907564      0
#> 810  -0.9126689366      0
#> 811  -1.4563535323      0
#> 812  -0.1282962865      0
#> 813  -1.4271578793      0
#> 814   1.0479436522      0
#> 815   0.3924590074      0
#> 816  -0.2464478148      0
#> 817  -0.4134770025      0
#> 818   0.6942038567      0
#> 819  -0.6138121684      0
#> 820   0.5740861053      0
#> 821  -0.1577769878      0
#> 822  -0.2093213371      0
#> 823   0.4336437770      0
#> 824  -0.7687008533      0
#> 825  -0.0961193959      0
#> 826   0.9758189303      0
#> 827   0.0469282127      0
#> 828   0.6557920064      0
#> 829   0.2918156467      0
#> 830  -2.4843528645      0
#> 831  -0.0742328737      0
#> 832  -1.0842662919      0
#> 833   0.2865582167      0
#> 834   0.2120047250      0
#> 835  -0.8454556428      0
#> 836   0.0903012130      0
#> 837   0.4601187854      0
#> 838  -0.3128533650      0
#> 839   1.2229642610      0
#> 840   0.2442107716      0
#> 841   1.2015719559      0
#> 842  -0.6833249262      0
#> 843  -0.9339563534      0
#> 844  -0.1808356110      0
#> 845  -0.5144363762      0
#> 846  -0.0736150875      0
#> 847  -1.3985912483      0
#> 848   0.3437355001      0
#> 849   0.1257133714      0
#> 850  -0.8788569782      0
#> 851  -0.6517016962      0
#> 852   1.3768378088      0
#> 853  -2.2598645831      0
#> 854   0.0784060552      0
#> 855  -0.1268977125      0
#> 856   0.2466547409      0
#> 857   0.3627476274      0
#> 858   0.0929235523      0
#> 859   0.1606449467      0
#> 860   0.5033903412      0
#> 861  -1.6034740119      0
#> 862  -0.4183249319      0
#> 863  -0.6075644388      0
#> 864  -0.0553759686      0
#> 865  -0.6051196189      0
#> 866   0.2482433808      0
#> 867  -1.1572643834      0
#> 868  -2.7243238933      0
#> 869   0.6305194453      0
#> 870   1.1960603001      0
#> 871   0.2556619341      0
#> 872  -0.4363396535      0
#> 873  -0.2054073677      0
#> 874  -0.6388224419      0
#> 875   0.1994870182      0
#> 876   0.1072184113      0
#> 877   0.0927599695      0
#> 878  -0.7955370354      0
#> 879  -0.4379346737      0
#> 880   1.5151188839      0
#> 881   0.0612189694      0
#> 882  -0.9392967588      0
#> 883  -0.0160809866      0
#> 884   1.3085011402      0
#> 885  -1.5897997971      0
#> 886  -0.0187145430      0
#> 887   0.7808213629      0
#> 888  -0.7771129704      0
#> 889   1.6582967804      0
#> 890   0.3992034314      0
#> 891   0.2191965952      0
#> 892   0.0271418302      0
#> 893   0.6318987607      0
#> 894   1.3465293930      0
#> 895  -2.3041832173      0
#> 896  -0.1458705698      0
#> 897  -0.0373291448      0
#> 898  -0.8370794594      0
#> 899  -0.8473266522      0
#> 900  -0.8666372294      0
#> 901   2.5039674945      0
#> 902   1.7239311061      0
#> 903  -0.5037874471      0
#> 904   1.1752087767      1
#> 905   0.3020134016      1
#> 906   1.0998088510      1
#> 907   0.5277956256      1
#> 908   1.5177246392      1
#> 909   0.3138368565      1
#> 910  -1.0051519375      1
#> 911   1.2360888412      1
#> 912  -0.1477456072      1
#> 913   0.1557018659      1
#> 914  -1.0945692270      1
#> 915  -0.5968518797      1
#> 916   0.2949806593      1
#> 917   2.0011969757      1
#> 918   0.0738890909      1
#> 919   0.9540153919      1
#> 920   1.2260314075      1
#> 921   0.3349655157      1
#> 922   1.1273139546      1
#> 923  -0.6220702106      1
#> 924   0.5529556459      1
#> 925   1.0108056911      1
#> 926   2.4891872460      1
#> 927   1.1769426911      1
#> 928   1.6670822012      1
#> 929   2.0054990282      1
#> 930   1.2567415660      1
#> 931   1.3526930620      1
#> 932   1.1561029907      1
#> 933   1.0662311015      1
#> 934   0.0055625964      1
#> 935   0.8896895616      1
#> 936  -0.1301619719      1
#> 937  -0.2856527731      1
#> 938   0.4438713161      1
#> 939  -0.3731067379      1
#> 940   1.2657099096      1
#> 941   1.2251225229      1
#> 942   0.7710288839      1
#> 943  -0.3173916326      1
#> 944   2.0110073145      1
#> 945  -1.1342059225      1
#> 946   0.2993734343      1
#> 947   1.4233702023      1
#> 948   1.7849800123      1
#> 949  -0.0977677873      1
#> 950  -0.9347189224      1
#> 951   0.6714877562      1
#> 952   0.8465239969      1
#> 953   1.7870056949      1
#> 954   0.5219977509      1
#> 955   1.9204709654      1
#> 956   0.9240550379      1
#> 957   1.9893099928      1
#> 958   1.1282965448      1
#> 959   1.9402135432      1
#> 960   0.0417870428      1
#> 961  -0.2547433849      1
#> 962  -0.3771884874      1
#> 963   1.4195234601      1
#> 964   0.0685267838      1
#> 965   0.3231055116      1
#> 966  -0.3001168359      1
#> 967  -0.3875902745      1
#> 968   0.0770143956      1
#> 969   0.1714166873      1
#> 970   1.1300172055      1
#> 971   1.6152987220      1
#> 972   0.0915788666      1
#> 973   1.6116328605      1
#> 974  -0.3260760197      1
#> 975   2.3086522139      1
#> 976   1.1477199983      1
#> 977   0.2050863465      1
#> 978   0.4591567245      1
#> 979   1.5334691905      1
#> 980  -0.0864714900      1
#> 981   2.1304423437      1
#> 982   0.0840946224      1
#> 983   1.2233718994      1
#> 984  -1.7202624755      1
#> 985   0.5924703922      1
#> 986  -0.7367305090      1
#> 987   1.4554006973      1
#> 988   0.5755443883      1
#> 989   0.3619374456      1
#> 990   0.0174275163      1
#> 991   0.7443699045      1
#> 992   3.2371483442      1
#> 993  -1.0461508791      1
#> 994   0.4973395624      1
#> 995   1.1877611841      1
#> 996   1.3512389415      1
#> 997   0.9958621547      1
#> 998   0.6217506628      1
#> 999   1.4153774196      1
#> 1000  0.9782483636      1