Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.188994312      0
#> 2    -1.405409176      0
#> 3    -0.572814620      0
#> 4     0.472216173      0
#> 5    -0.250692194      0
#> 6     0.309944946      0
#> 7     0.173517871      0
#> 8    -1.770423937      0
#> 9     0.581654871      0
#> 10    0.456215425      0
#> 11   -0.423816793      0
#> 12   -0.167480902      0
#> 13    1.261616301      0
#> 14   -1.582916171      0
#> 15   -0.157476155      0
#> 16    0.709645635      0
#> 17    1.069431444      0
#> 18    0.467787627      0
#> 19   -1.023280314      0
#> 20    1.044124436      0
#> 21    0.295924697      0
#> 22    0.633852880      0
#> 23    0.904699973      0
#> 24    0.742010834      0
#> 25   -2.068841738      0
#> 26    0.200451603      0
#> 27    0.687289854      0
#> 28   -0.022153822      0
#> 29    2.050621349      0
#> 30   -0.788288028      0
#> 31    0.930927108      0
#> 32   -0.077660508      0
#> 33   -0.147723128      0
#> 34    0.444132642      0
#> 35    1.190317162      0
#> 36   -0.393521237      0
#> 37   -0.596200290      0
#> 38   -0.072815061      0
#> 39   -0.389317201      0
#> 40    1.283565894      0
#> 41    1.504328379      0
#> 42    1.514820356      0
#> 43   -0.374561598      0
#> 44    0.080429790      0
#> 45   -0.814990657      0
#> 46   -0.679523982      0
#> 47    0.884598691      0
#> 48   -0.827427252      0
#> 49    1.505631553      0
#> 50   -0.505831230      0
#> 51    0.405410584      0
#> 52    0.753408475      0
#> 53    2.419411165      0
#> 54   -0.003600887      0
#> 55    0.622013036      0
#> 56   -0.048061130      0
#> 57   -1.134269917      0
#> 58   -1.452228741      0
#> 59    0.514903461      0
#> 60   -0.483905395      0
#> 61   -1.072079568      0
#> 62    0.509955027      0
#> 63    1.039731158      0
#> 64    0.163975842      0
#> 65   -0.249140461      0
#> 66   -1.102985343      0
#> 67    0.974113740      0
#> 68    1.763281689      0
#> 69    0.476547403      0
#> 70   -0.957824757      0
#> 71   -0.352392428      0
#> 72    0.448257208      0
#> 73    0.388251890      0
#> 74   -0.509940002      0
#> 75   -2.009277791      0
#> 76    0.134836879      0
#> 77   -0.626391342      0
#> 78   -3.081760245      0
#> 79   -0.005945460      0
#> 80    1.207485353      0
#> 81    0.608084930      0
#> 82   -0.875578852      0
#> 83   -0.186168999      0
#> 84   -0.583767122      0
#> 85    0.599492748      0
#> 86    1.864065689      0
#> 87   -0.178293653      0
#> 88    1.251189594      0
#> 89    1.011289464      0
#> 90   -1.351133229      0
#> 91    1.926694709      0
#> 92   -1.536249889      0
#> 93    0.849087695      0
#> 94   -0.197663616      0
#> 95    0.871938106      0
#> 96   -1.505158515      0
#> 97   -0.319318092      0
#> 98   -0.259374839      0
#> 99   -1.157820109      0
#> 100  -0.642044838      0
#> 101   0.055480908      0
#> 102  -0.736845692      0
#> 103  -1.152151561      0
#> 104  -1.285094206      0
#> 105  -0.788320718      0
#> 106  -2.542117881      0
#> 107  -1.079636277      0
#> 108  -0.158228442      0
#> 109  -0.277102493      0
#> 110  -0.666351974      0
#> 111   2.213974271      0
#> 112   0.031650351      0
#> 113  -0.837142616      0
#> 114   0.192888774      0
#> 115   1.233081761      0
#> 116  -2.045690308      0
#> 117   1.381559675      0
#> 118   0.669916707      0
#> 119   0.974087594      0
#> 120   1.458663470      0
#> 121   0.295693254      0
#> 122   0.422608112      0
#> 123   0.114301033      0
#> 124   0.353061849      0
#> 125   1.145555215      0
#> 126  -0.910950955      0
#> 127   0.098502262      0
#> 128  -0.221371539      0
#> 129   0.915697227      0
#> 130   1.207773451      0
#> 131   1.687215218      0
#> 132  -0.278715834      0
#> 133  -2.656952698      0
#> 134  -0.318103154      0
#> 135  -0.030471781      0
#> 136   0.699969873      0
#> 137  -0.331901523      0
#> 138   0.910724587      0
#> 139  -0.527043319      0
#> 140   0.742792913      0
#> 141  -0.323997004      0
#> 142  -0.425082736      0
#> 143  -0.006287534      0
#> 144  -1.453342576      0
#> 145   0.415899022      0
#> 146   0.571845729      0
#> 147   0.110320176      0
#> 148  -1.353597592      0
#> 149  -0.313776923      0
#> 150  -1.306988609      0
#> 151   0.070006982      0
#> 152  -0.447355271      0
#> 153  -0.233029026      0
#> 154  -0.101912761      0
#> 155  -0.351070699      0
#> 156  -0.443192915      0
#> 157  -0.408761817      0
#> 158  -0.168338423      0
#> 159  -2.206429144      0
#> 160   0.066795523      0
#> 161   0.275807267      0
#> 162   0.143860069      0
#> 163   1.051198431      0
#> 164   1.828522699      0
#> 165   0.186141822      0
#> 166  -0.392866115      0
#> 167  -1.219861625      0
#> 168  -0.321635837      0
#> 169   0.543426706      0
#> 170   2.020794142      0
#> 171  -1.851493171      0
#> 172   0.226613611      0
#> 173  -1.865646394      0
#> 174  -2.262643911      0
#> 175   0.593014851      0
#> 176  -0.309903922      0
#> 177  -0.840196908      0
#> 178  -0.107347372      0
#> 179   1.267849791      0
#> 180   1.184521310      0
#> 181   2.820388952      0
#> 182  -0.012026394      0
#> 183   1.333276410      0
#> 184  -0.388374444      0
#> 185  -0.100205355      0
#> 186  -1.669457186      0
#> 187  -0.076044664      0
#> 188   0.931600867      0
#> 189  -0.279878941      0
#> 190  -0.360190595      0
#> 191   1.428034941      0
#> 192  -0.091371406      0
#> 193  -0.274465608      0
#> 194   2.098290521      0
#> 195  -1.174269823      0
#> 196   0.466812150      0
#> 197  -0.359655801      0
#> 198  -0.751201269      0
#> 199  -0.719408906      0
#> 200  -2.161327110      0
#> 201   1.106886345      0
#> 202  -1.007770533      0
#> 203   1.770085491      0
#> 204  -1.577004973      0
#> 205   0.428077626      0
#> 206   0.744278785      0
#> 207   0.192424461      0
#> 208  -0.755702922      0
#> 209  -1.314854361      0
#> 210   0.560426076      0
#> 211   0.189664430      0
#> 212   1.723712137      0
#> 213   0.824845816      0
#> 214   0.432288472      0
#> 215   0.715533907      0
#> 216   0.664781836      0
#> 217  -0.406579048      0
#> 218   0.295566981      0
#> 219  -0.415253061      0
#> 220  -0.989655263      0
#> 221   1.477024183      0
#> 222   0.701930468      0
#> 223  -0.603651543      0
#> 224   1.154153935      0
#> 225  -0.075797457      0
#> 226   1.652122980      0
#> 227   0.901304722      0
#> 228   0.762758894      0
#> 229  -0.070374577      0
#> 230  -0.191014464      0
#> 231  -0.188713374      0
#> 232   0.172103871      0
#> 233   0.610281384      0
#> 234  -0.548504649      0
#> 235   1.451812026      0
#> 236   0.341087521      0
#> 237   1.343224411      0
#> 238  -0.292052580      0
#> 239   0.993429245      0
#> 240   1.023375921      0
#> 241  -0.453625587      0
#> 242  -1.213968411      0
#> 243   2.078868122      0
#> 244   1.856500551      0
#> 245  -0.304394166      0
#> 246   0.428378958      0
#> 247   0.582442962      0
#> 248  -0.032581214      0
#> 249  -1.323204913      0
#> 250   0.584312452      0
#> 251   0.953950669      0
#> 252   1.415029496      0
#> 253  -1.206641451      0
#> 254  -0.855166042      0
#> 255  -0.051464775      0
#> 256  -0.556981079      0
#> 257   0.101875363      0
#> 258   0.024569113      0
#> 259  -1.343903325      0
#> 260  -0.322408613      0
#> 261   0.844661934      0
#> 262  -0.669956269      0
#> 263   0.826400918      0
#> 264   0.616642111      0
#> 265   0.889410403      0
#> 266   0.574027706      0
#> 267  -0.586479135      0
#> 268   0.637177453      0
#> 269   1.595628466      0
#> 270   2.218745674      0
#> 271   0.255192881      0
#> 272   0.818703322      0
#> 273  -0.033426665      0
#> 274  -0.764276952      0
#> 275  -0.051076451      0
#> 276   1.181333425      0
#> 277  -0.999615210      0
#> 278   0.276671412      0
#> 279   1.620402275      0
#> 280   1.278654886      0
#> 281  -0.433036387      0
#> 282  -1.717061181      0
#> 283   0.905125826      0
#> 284  -0.194548182      0
#> 285  -1.519616262      0
#> 286  -1.208759694      0
#> 287  -0.202867230      0
#> 288   0.592925180      0
#> 289   0.524144120      0
#> 290  -0.271100333      0
#> 291  -0.283428794      0
#> 292  -1.005455671      0
#> 293   0.422593378      0
#> 294   2.059544969      0
#> 295  -0.287855484      0
#> 296   0.381579302      0
#> 297   1.456806257      0
#> 298  -1.209119138      0
#> 299   0.314899343      0
#> 300  -0.650447522      0
#> 301   0.901691407      0
#> 302  -0.001932576      0
#> 303   0.211897626      0
#> 304   0.240393612      0
#> 305  -0.051471888      0
#> 306  -0.617335395      0
#> 307   0.727606747      0
#> 308   0.195759080      0
#> 309   0.684931169      0
#> 310  -0.778404679      0
#> 311   1.296764133      0
#> 312  -0.382615645      0
#> 313   1.037501089      0
#> 314   1.010503920      0
#> 315  -0.133119749      0
#> 316   1.186319252      0
#> 317  -0.276421818      0
#> 318   0.346637666      0
#> 319   1.406425636      0
#> 320   0.264182472      0
#> 321  -0.682031093      0
#> 322  -0.184227351      0
#> 323  -0.003339807      0
#> 324  -0.719735120      0
#> 325   0.387554121      0
#> 326  -0.834674179      0
#> 327  -0.960157917      0
#> 328   1.786317475      0
#> 329  -0.800692038      0
#> 330  -1.033986662      0
#> 331   0.122587138      0
#> 332  -0.948440670      0
#> 333   0.253167073      0
#> 334  -0.274167052      0
#> 335  -0.606232949      0
#> 336   0.208270485      0
#> 337  -1.740449903      0
#> 338   1.625017965      0
#> 339  -0.109303381      0
#> 340   0.770563376      0
#> 341   0.669422707      0
#> 342   0.520724331      0
#> 343  -1.007760942      0
#> 344  -1.030572477      0
#> 345   1.117670526      0
#> 346  -0.069879206      0
#> 347  -0.163350952      0
#> 348   0.835040073      0
#> 349  -0.082634366      0
#> 350  -0.001771974      0
#> 351  -0.082210091      0
#> 352  -1.210168155      0
#> 353   0.610446999      0
#> 354   0.150483761      0
#> 355  -0.136295039      0
#> 356   0.184671293      0
#> 357   0.283118724      0
#> 358   0.182929183      0
#> 359   0.306565808      0
#> 360   2.523185712      0
#> 361   1.079613054      0
#> 362   0.786171112      0
#> 363   2.340424039      0
#> 364   0.867757185      0
#> 365   1.104602785      0
#> 366  -0.264303789      0
#> 367  -0.093930301      0
#> 368   0.612124595      0
#> 369  -0.070486730      0
#> 370  -1.423286424      0
#> 371   0.479929370      0
#> 372   1.624991322      0
#> 373   0.168232003      0
#> 374   1.823348183      0
#> 375   1.246936093      0
#> 376  -0.051052364      0
#> 377  -0.205791457      0
#> 378   0.883407195      0
#> 379  -0.809094973      0
#> 380  -0.709475296      0
#> 381   1.500362284      0
#> 382  -0.404028608      0
#> 383   0.096062088      0
#> 384  -0.299437431      0
#> 385   0.176950987      0
#> 386   1.057304244      0
#> 387   0.475116552      0
#> 388   0.457193060      0
#> 389  -0.927355316      0
#> 390   1.113187902      0
#> 391  -0.632323221      0
#> 392   1.358579561      0
#> 393   0.420492722      0
#> 394  -0.184289244      0
#> 395  -0.162259414      0
#> 396   0.525896944      0
#> 397  -1.067559270      0
#> 398  -0.145976365      0
#> 399   1.140695679      0
#> 400   0.120874313      0
#> 401  -1.239496817      0
#> 402   1.128928210      0
#> 403   0.856767412      0
#> 404  -1.402162350      0
#> 405  -0.860813813      0
#> 406  -2.358952780      0
#> 407   1.246871848      0
#> 408  -1.711782020      0
#> 409   2.545600185      0
#> 410   1.308107822      0
#> 411   0.628464142      0
#> 412  -1.797783505      0
#> 413  -0.238636339      0
#> 414   1.441690228      0
#> 415  -0.057768948      0
#> 416   0.175632763      0
#> 417  -0.288457231      0
#> 418  -1.940324701      0
#> 419  -0.446982839      0
#> 420  -0.129398605      0
#> 421   0.737104402      0
#> 422   0.458390848      0
#> 423   0.435182557      0
#> 424  -0.408836179      0
#> 425   0.926547969      0
#> 426  -0.068658933      0
#> 427   0.364802496      0
#> 428   0.520004857      0
#> 429  -0.751080510      0
#> 430   0.521077441      0
#> 431  -0.795022148      0
#> 432  -0.306840815      0
#> 433  -0.346842748      0
#> 434   0.269977941      0
#> 435  -1.341418688      0
#> 436  -0.511318923      0
#> 437  -0.466792548      0
#> 438   0.787660012      0
#> 439  -0.275830294      0
#> 440   0.180774228      0
#> 441  -0.055417837      0
#> 442  -1.889230613      0
#> 443  -1.057635316      0
#> 444   0.266203926      0
#> 445   1.067535372      0
#> 446  -1.131732776      0
#> 447  -1.587359345      0
#> 448  -0.635425451      0
#> 449   0.104434595      0
#> 450   0.008249698      0
#> 451  -0.198889441      0
#> 452   0.838290431      0
#> 453   0.052404190      0
#> 454  -0.228067476      0
#> 455   0.126126114      0
#> 456  -0.201351728      0
#> 457  -0.183384868      0
#> 458   0.302189063      0
#> 459  -0.938008000      0
#> 460  -0.275548528      0
#> 461  -0.146232166      0
#> 462  -0.955463784      0
#> 463  -0.667874406      0
#> 464   0.194428855      0
#> 465  -0.286854934      0
#> 466   1.922503456      0
#> 467   0.923492952      0
#> 468  -1.799520267      0
#> 469   0.605333213      0
#> 470   0.058892811      0
#> 471   1.625469505      0
#> 472   0.121658986      0
#> 473   0.476794280      0
#> 474   0.375175359      0
#> 475   1.602088275      0
#> 476   0.150832475      0
#> 477  -0.695472361      0
#> 478  -1.336768296      0
#> 479  -0.240069626      0
#> 480   0.995235615      0
#> 481  -0.614970407      0
#> 482   1.049204075      0
#> 483   0.645731014      0
#> 484   0.492343874      0
#> 485   0.176644921      0
#> 486  -0.038823746      0
#> 487  -0.228099438      0
#> 488   0.029579918      0
#> 489   1.026182239      0
#> 490  -0.150890465      0
#> 491   1.543965921      0
#> 492  -0.863789920      0
#> 493  -2.101647303      0
#> 494   1.136516862      0
#> 495   0.384826087      0
#> 496  -0.019215149      0
#> 497   0.728718868      0
#> 498  -0.058708902      0
#> 499  -0.700434517      0
#> 500   0.479202343      0
#> 501   0.234393181      0
#> 502   0.663670831      0
#> 503   0.988666280      0
#> 504   0.521122536      0
#> 505   1.592676151      0
#> 506  -0.433477276      0
#> 507   0.301778295      0
#> 508   0.911790820      0
#> 509  -0.493173173      0
#> 510   1.533861737      0
#> 511  -0.767581210      0
#> 512   1.806643768      0
#> 513   0.087531827      0
#> 514   1.287471789      0
#> 515  -1.332069772      0
#> 516  -0.950380168      0
#> 517  -1.666279725      0
#> 518  -2.395264763      0
#> 519   0.047118164      0
#> 520   2.910887866      0
#> 521  -1.587644827      0
#> 522   1.295656382      0
#> 523  -1.401665946      0
#> 524   0.117281213      0
#> 525  -2.209698232      0
#> 526  -1.175746675      0
#> 527  -0.181500063      0
#> 528  -1.334588498      0
#> 529   0.465728066      0
#> 530   1.312925780      0
#> 531  -0.136382578      0
#> 532  -0.358332316      0
#> 533  -0.737934653      0
#> 534   0.463534589      0
#> 535   0.164509900      0
#> 536   0.044750623      0
#> 537   1.311265522      0
#> 538   0.820486614      0
#> 539   1.070168465      0
#> 540   1.326803919      0
#> 541   1.565382615      0
#> 542   0.581817541      0
#> 543  -0.409452530      0
#> 544   1.104559635      0
#> 545   0.321689888      0
#> 546  -0.834655935      0
#> 547   0.598372338      0
#> 548  -0.214951474      0
#> 549  -2.050397753      0
#> 550   0.659222263      0
#> 551  -0.885318654      0
#> 552   0.018481203      0
#> 553  -0.897146073      0
#> 554  -0.711877357      0
#> 555  -1.176999605      0
#> 556   1.275104907      0
#> 557   0.466457765      0
#> 558  -0.264937483      0
#> 559  -1.565673553      0
#> 560  -1.458770139      0
#> 561  -0.415911018      0
#> 562  -0.930565737      0
#> 563  -0.947772780      0
#> 564   0.032755125      0
#> 565   0.909517697      0
#> 566   0.551389779      0
#> 567  -0.552114024      0
#> 568   0.110877574      0
#> 569  -0.510978524      0
#> 570  -1.622126532      0
#> 571   0.148670134      0
#> 572  -1.290591020      0
#> 573   0.925432683      0
#> 574   0.417529390      0
#> 575  -1.256255295      0
#> 576  -2.333944575      0
#> 577  -0.549107755      0
#> 578  -0.120644818      0
#> 579   0.320695980      0
#> 580   0.733426056      0
#> 581   0.106299683      0
#> 582   0.319194316      0
#> 583   1.099821131      0
#> 584   0.177218977      0
#> 585   1.441744605      0
#> 586   1.107210899      0
#> 587  -1.879993663      0
#> 588   0.294474210      0
#> 589   0.183569642      0
#> 590   1.533340260      0
#> 591   0.435348861      0
#> 592  -0.601488913      0
#> 593  -0.410478956      0
#> 594   0.068016234      0
#> 595  -0.106849190      0
#> 596   1.151375976      0
#> 597   1.243545203      0
#> 598   0.158705238      0
#> 599   0.174485911      0
#> 600  -0.208622350      0
#> 601   0.777915711      0
#> 602  -2.199673640      0
#> 603   0.174228680      0
#> 604  -0.958441862      0
#> 605  -0.510352402      0
#> 606  -1.522420888      0
#> 607  -1.552231247      0
#> 608  -0.564012270      0
#> 609   0.321925802      0
#> 610  -0.133323821      0
#> 611  -0.106777543      0
#> 612   0.417088652      0
#> 613   0.338787284      0
#> 614  -1.377154285      0
#> 615  -0.931849220      0
#> 616   0.009410711      0
#> 617  -0.350640566      0
#> 618   0.173456585      0
#> 619   0.567046358      0
#> 620   0.989046251      0
#> 621   1.909716841      0
#> 622  -0.337423998      0
#> 623  -0.438047553      0
#> 624   0.002219439      0
#> 625   1.416352911      0
#> 626   0.071612566      0
#> 627  -0.310052951      0
#> 628   1.780015197      0
#> 629   1.000120260      0
#> 630  -0.806981444      0
#> 631  -2.171950553      0
#> 632   1.677381632      0
#> 633   0.464836067      0
#> 634  -0.474903341      0
#> 635  -0.186033625      0
#> 636   0.421395042      0
#> 637  -0.136352815      0
#> 638   0.207841552      0
#> 639   1.843155456      0
#> 640  -0.727573155      0
#> 641   0.488913523      0
#> 642   0.163117813      0
#> 643   1.720138547      0
#> 644  -0.981409350      0
#> 645  -0.261110209      0
#> 646   0.088361515      0
#> 647   0.150232658      0
#> 648  -0.538837529      0
#> 649   0.316713750      0
#> 650   0.573812084      0
#> 651   1.540816352      0
#> 652  -0.928842393      0
#> 653  -1.453665218      0
#> 654  -0.296910678      0
#> 655   0.555679891      0
#> 656  -0.257341009      0
#> 657   0.168362271      0
#> 658  -0.087932701      0
#> 659  -0.112765978      0
#> 660  -1.216867677      0
#> 661  -0.862581914      0
#> 662   0.789725070      0
#> 663   0.893863857      0
#> 664  -0.094680002      0
#> 665  -1.024553109      0
#> 666   1.393294265      0
#> 667  -1.924588261      0
#> 668   0.615846818      0
#> 669   0.721094653      0
#> 670  -0.362754318      0
#> 671  -0.709812800      0
#> 672  -0.097761058      0
#> 673   0.542897434      0
#> 674   0.389534215      0
#> 675  -1.228961507      0
#> 676  -0.326870767      0
#> 677   0.904463821      0
#> 678  -1.565448574      0
#> 679  -1.702339175      0
#> 680  -1.341000316      0
#> 681   0.635143595      0
#> 682  -0.825917237      0
#> 683   0.189679050      0
#> 684   0.180938725      0
#> 685   0.439058507      0
#> 686   0.887100035      0
#> 687   0.394984381      0
#> 688   0.436418031      0
#> 689   0.234769448      0
#> 690  -0.613908538      0
#> 691  -2.934343424      0
#> 692   0.427458142      0
#> 693   0.560850343      0
#> 694  -0.585268430      0
#> 695  -1.596192866      0
#> 696   0.383404953      0
#> 697   0.106545467      0
#> 698  -0.667895694      0
#> 699  -1.582782474      0
#> 700  -1.154798257      0
#> 701   1.000382046      0
#> 702   0.222853467      0
#> 703  -0.993629176      0
#> 704  -1.174798484      0
#> 705  -0.884467008      0
#> 706  -0.554705366      0
#> 707   1.080412005      0
#> 708  -0.611101153      0
#> 709  -1.579031139      0
#> 710   0.570876321      0
#> 711   0.669563614      0
#> 712   0.320919912      0
#> 713   1.406132696      0
#> 714   0.192391173      0
#> 715   0.096811055      0
#> 716   0.913105831      0
#> 717  -0.918096015      0
#> 718   1.673332455      0
#> 719  -0.478600860      0
#> 720   0.363138602      0
#> 721  -0.053011505      0
#> 722   0.070870579      0
#> 723  -0.743053629      0
#> 724   0.327446382      0
#> 725   1.186731883      0
#> 726   1.585860348      0
#> 727  -0.776822508      0
#> 728  -1.518060927      0
#> 729   0.833570688      0
#> 730  -0.458932291      0
#> 731  -0.700278825      0
#> 732   0.614668656      0
#> 733  -0.530527166      0
#> 734  -1.922381719      0
#> 735   0.233717434      0
#> 736  -0.255232865      0
#> 737  -1.325389763      0
#> 738   0.538051354      0
#> 739  -1.086188841      0
#> 740   0.719841661      0
#> 741  -0.449686035      0
#> 742  -1.473303700      0
#> 743   0.118608464      0
#> 744   1.469660054      0
#> 745  -0.776392228      0
#> 746  -1.581456757      0
#> 747  -0.799051263      0
#> 748   1.134060406      0
#> 749  -0.683716640      0
#> 750  -1.765389274      0
#> 751  -0.355587946      0
#> 752  -1.016379998      0
#> 753  -0.180141392      0
#> 754   0.774086032      0
#> 755   1.385777935      0
#> 756  -0.337693029      0
#> 757  -0.598181285      0
#> 758  -1.977664609      0
#> 759  -0.590930694      0
#> 760   1.735994296      0
#> 761   0.408547826      0
#> 762   1.159545623      0
#> 763  -0.198026912      0
#> 764  -0.603036443      0
#> 765  -1.079996300      0
#> 766   0.541679997      0
#> 767  -0.187240938      0
#> 768   1.103578117      0
#> 769   0.498094138      0
#> 770  -1.395895308      0
#> 771   1.431894388      0
#> 772   2.057436314      0
#> 773   0.251587748      0
#> 774   0.169106752      0
#> 775  -0.409927988      0
#> 776   1.385878066      0
#> 777   0.666122026      0
#> 778   0.064030402      0
#> 779  -0.267179691      0
#> 780   0.810817810      0
#> 781  -1.107291216      0
#> 782  -0.606732094      0
#> 783   0.288155827      0
#> 784  -1.306889440      0
#> 785  -0.022283460      0
#> 786   1.594831308      0
#> 787  -0.402752392      0
#> 788  -1.561529045      0
#> 789   1.968816328      0
#> 790   1.174119080      0
#> 791  -0.489911652      0
#> 792  -0.329888458      0
#> 793   1.607383247      0
#> 794   1.513454869      0
#> 795  -0.295638205      0
#> 796  -0.768980085      0
#> 797  -0.990979396      0
#> 798  -0.874455366      0
#> 799   1.148190246      0
#> 800  -1.480134950      0
#> 801   0.540838143      0
#> 802   0.536464393      0
#> 803  -0.946341311      0
#> 804  -0.563176912      0
#> 805  -1.119458983      0
#> 806   0.909760218      0
#> 807  -1.305563829      0
#> 808  -0.434620787      0
#> 809   0.199893758      0
#> 810   0.443443523      0
#> 811  -0.086903130      0
#> 812   1.693112472      0
#> 813  -0.776405626      0
#> 814   0.210895350      0
#> 815   1.319487779      0
#> 816  -2.667474538      0
#> 817   0.331447582      0
#> 818  -1.334720385      0
#> 819   1.008840469      0
#> 820  -2.444513585      0
#> 821  -1.113181465      0
#> 822   0.976397271      0
#> 823  -1.592676802      0
#> 824  -0.181102735      0
#> 825   0.279268313      0
#> 826   0.978153403      0
#> 827   1.951860821      0
#> 828   2.075866826      0
#> 829   0.236283015      0
#> 830  -1.401706475      0
#> 831   0.181964954      0
#> 832  -2.321587062      0
#> 833   0.127183317      0
#> 834   1.216997768      0
#> 835   1.147288888      0
#> 836   0.668836006      0
#> 837  -1.519802404      0
#> 838  -0.415642707      0
#> 839   0.250458200      0
#> 840  -2.304045537      0
#> 841  -2.473636716      0
#> 842  -0.140728693      0
#> 843  -0.422989751      0
#> 844   0.035801096      0
#> 845   1.836542225      0
#> 846  -0.579445195      0
#> 847  -0.055342944      0
#> 848   0.206426744      0
#> 849  -0.108568566      0
#> 850  -0.944874794      0
#> 851   0.977482698      0
#> 852  -1.097746412      0
#> 853  -0.700746276      0
#> 854   0.447210461      0
#> 855   0.404308073      0
#> 856  -0.147656362      0
#> 857  -0.044040122      0
#> 858   0.178809612      0
#> 859   0.687855470      0
#> 860   1.826204593      0
#> 861   0.029134575      0
#> 862   0.494594113      0
#> 863   0.820799642      0
#> 864  -1.835555620      0
#> 865   0.192014887      0
#> 866  -0.051282702      0
#> 867  -0.568104531      0
#> 868  -1.229395331      0
#> 869   0.512666259      0
#> 870  -0.922867813      0
#> 871  -0.546753575      0
#> 872   1.335322936      0
#> 873  -0.349009561      0
#> 874  -0.657850365      0
#> 875   0.084272799      0
#> 876  -0.203681478      0
#> 877   0.703856899      0
#> 878   1.167277796      0
#> 879   0.814123566      0
#> 880  -0.181985956      0
#> 881   0.629264548      0
#> 882  -1.115609815      0
#> 883  -0.182454712      0
#> 884   0.538468851      0
#> 885  -0.153156694      0
#> 886   0.450155087      0
#> 887  -0.395768540      0
#> 888   1.035219902      0
#> 889  -0.180289110      0
#> 890  -1.283202409      0
#> 891   0.228886278      0
#> 892   0.704757891      0
#> 893   0.241297006      0
#> 894   1.736438192      0
#> 895  -0.233166848      0
#> 896  -0.042914219      0
#> 897  -0.613354619      0
#> 898  -0.320169518      0
#> 899  -0.054955850      0
#> 900   1.038342086      0
#> 901   0.348545877      0
#> 902  -0.996390177      0
#> 903  -0.423923170      0
#> 904   1.637437826      0
#> 905   0.185338222      0
#> 906   0.395751111      0
#> 907  -0.328912259      0
#> 908   2.283322751      0
#> 909  -0.714634811      0
#> 910  -0.356969485      0
#> 911  -0.851650139      0
#> 912   1.168612672      0
#> 913  -0.969910948      1
#> 914   0.986862161      1
#> 915   1.572155654      1
#> 916  -0.577043230      1
#> 917  -0.482783844      1
#> 918   0.923029911      1
#> 919   0.402229951      1
#> 920   2.954635411      1
#> 921  -0.289918201      1
#> 922   1.400612313      1
#> 923   0.509471227      1
#> 924   2.970349270      1
#> 925  -0.968842984      1
#> 926   1.347993441      1
#> 927   0.751689866      1
#> 928   0.555328056      1
#> 929  -0.089739546      1
#> 930   0.970653537      1
#> 931   0.911132122      1
#> 932   0.352783563      1
#> 933   0.686026202      1
#> 934  -0.738269566      1
#> 935   0.753293497      1
#> 936  -1.779939476      1
#> 937   0.083703329      1
#> 938   0.481017696      1
#> 939  -0.571354689      1
#> 940   0.757114520      1
#> 941   0.688914496      1
#> 942   0.429482865      1
#> 943  -0.526477758      1
#> 944   0.333752636      1
#> 945   1.522632120      1
#> 946   0.651390388      1
#> 947   0.759376457      1
#> 948   0.731610917      1
#> 949   1.098966784      1
#> 950   0.327052456      1
#> 951   0.573594960      1
#> 952   0.433512464      1
#> 953   0.389977625      1
#> 954   1.330931355      1
#> 955   0.662750487      1
#> 956   1.627345084      1
#> 957  -0.027898706      1
#> 958   0.292841441      1
#> 959   0.358750886      1
#> 960  -0.158580068      1
#> 961  -1.631055409      1
#> 962  -0.415558391      1
#> 963   2.194105211      1
#> 964  -1.501535594      1
#> 965   1.706521754      1
#> 966  -1.212609283      1
#> 967  -0.062404960      1
#> 968   1.397360449      1
#> 969   0.803222534      1
#> 970   3.414652051      1
#> 971   1.747025870      1
#> 972   1.604730318      1
#> 973   1.065624555      1
#> 974   0.204817880      1
#> 975   4.061958279      1
#> 976   1.157696644      1
#> 977   1.287993001      1
#> 978  -1.058058975      1
#> 979   0.866562335      1
#> 980   0.290857856      1
#> 981   0.703560126      1
#> 982   0.928083811      1
#> 983   1.597509030      1
#> 984  -0.553769468      1
#> 985   1.641295208      1
#> 986  -0.071831828      1
#> 987   2.727924395      1
#> 988  -1.051009397      1
#> 989  -0.385800732      1
#> 990  -0.436886168      1
#> 991  -0.009351262      1
#> 992   0.710122712      1
#> 993  -1.559828877      1
#> 994   1.403479835      1
#> 995   1.554563029      1
#> 996   0.691040368      1
#> 997   0.944606490      1
#> 998   0.760862016      1
#> 999  -0.337759610      1
#> 1000  0.722475222      1