Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -1.284621047      0
#> 2     0.249323365      0
#> 3     1.051198036      0
#> 4     1.279660477      0
#> 5    -1.106057036      0
#> 6    -0.089511423      0
#> 7    -0.789836021      0
#> 8     0.239402211      0
#> 9     1.780397525      0
#> 10    1.092450648      0
#> 11    1.028064707      0
#> 12    0.079025615      0
#> 13   -0.099627193      0
#> 14    0.397580285      0
#> 15   -0.348835165      0
#> 16    0.650068734      0
#> 17   -0.166657193      0
#> 18    1.238338226      0
#> 19    2.354397352      0
#> 20   -0.716188094      0
#> 21    0.285009005      0
#> 22    0.713159230      0
#> 23    0.053520336      0
#> 24   -0.401648046      0
#> 25   -2.109783518      0
#> 26   -1.092630156      0
#> 27   -0.415297587      0
#> 28    0.160680610      0
#> 29    1.090648516      0
#> 30   -0.029801457      0
#> 31    1.290893355      0
#> 32    0.034345862      0
#> 33    0.858079646      0
#> 34    0.582613851      0
#> 35    0.068456157      0
#> 36    1.613853481      0
#> 37    1.136524974      0
#> 38    1.084624531      0
#> 39   -0.766367692      0
#> 40    0.272620366      0
#> 41   -0.062554764      0
#> 42   -1.058734399      0
#> 43    1.849779217      0
#> 44    0.562283951      0
#> 45   -0.384701147      0
#> 46    1.241103145      0
#> 47    1.570561694      0
#> 48   -1.918603018      0
#> 49    0.817706562      0
#> 50    1.804152476      0
#> 51   -0.329829664      0
#> 52    0.459983617      0
#> 53   -1.177302344      0
#> 54   -1.091673471      0
#> 55    0.699897132      0
#> 56    0.073910204      0
#> 57   -0.884152255      0
#> 58    0.441273700      0
#> 59    0.445295142      0
#> 60   -0.481130414      0
#> 61   -0.433519356      0
#> 62   -0.295087757      0
#> 63   -0.049293551      0
#> 64   -0.804402549      0
#> 65   -0.398845219      0
#> 66   -0.356792544      0
#> 67   -0.473895658      0
#> 68    0.665126763      0
#> 69   -1.662196875      0
#> 70   -0.478359299      0
#> 71    0.990427606      0
#> 72   -0.118508784      0
#> 73   -0.587628905      0
#> 74    0.874677631      0
#> 75   -0.915125062      0
#> 76   -0.814105913      0
#> 77    1.896083080      0
#> 78    1.252477196      0
#> 79   -0.643301804      0
#> 80   -0.442918876      0
#> 81   -2.010540681      0
#> 82   -0.608482725      0
#> 83   -0.923568336      0
#> 84    0.087772782      0
#> 85    0.592972957      0
#> 86    1.269033166      0
#> 87    1.532539429      0
#> 88   -0.919344063      0
#> 89    1.840950916      0
#> 90   -1.027168370      0
#> 91    1.219499395      0
#> 92    0.671537703      0
#> 93    0.330462710      0
#> 94   -1.076650906      0
#> 95   -0.163861365      0
#> 96   -0.019951024      0
#> 97    0.530868086      0
#> 98    2.111472132      0
#> 99   -1.104749423      0
#> 100  -0.576649910      0
#> 101  -0.683864971      0
#> 102  -0.209121150      0
#> 103   1.546929276      0
#> 104  -0.584758924      0
#> 105   1.534558167      0
#> 106  -2.259309287      0
#> 107   2.159795212      0
#> 108  -1.640331895      0
#> 109  -0.915555440      0
#> 110  -0.151021416      0
#> 111  -0.993722200      0
#> 112   0.258205826      0
#> 113   0.398068107      0
#> 114   0.669779233      0
#> 115  -0.376953700      0
#> 116  -0.826945056      0
#> 117   0.065026295      0
#> 118   0.836261725      0
#> 119   0.959913756      0
#> 120  -1.176231934      0
#> 121  -2.072384486      0
#> 122   0.240890444      0
#> 123   0.248164942      0
#> 124   0.038311939      0
#> 125   0.192189563      0
#> 126   1.168568671      0
#> 127  -1.603847099      0
#> 128  -0.405320037      0
#> 129   0.572745608      0
#> 130   0.879054516      0
#> 131  -1.282418439      0
#> 132  -1.462610482      0
#> 133  -0.416755187      0
#> 134  -0.887921855      0
#> 135   0.155688182      0
#> 136   0.159555826      0
#> 137  -0.574068095      0
#> 138  -0.442297721      0
#> 139  -1.790976847      0
#> 140   3.242107520      0
#> 141   0.302457427      0
#> 142  -0.331817694      0
#> 143   0.692883983      0
#> 144   0.130205752      0
#> 145   0.754534719      0
#> 146   0.505726357      0
#> 147   1.483856650      0
#> 148  -0.109431387      0
#> 149   0.740042842      0
#> 150   0.644239218      0
#> 151  -1.524131624      0
#> 152  -0.190271882      0
#> 153   0.104509781      0
#> 154  -1.169051708      0
#> 155  -0.910292441      0
#> 156   1.084425930      0
#> 157   0.223866781      0
#> 158   0.424055631      0
#> 159   1.142002143      0
#> 160  -0.848627676      0
#> 161  -0.885941061      0
#> 162   1.352188871      0
#> 163   0.212200177      0
#> 164  -0.322019755      0
#> 165  -1.296246025      0
#> 166   0.122224381      0
#> 167   0.984287738      0
#> 168   0.091615678      0
#> 169   0.104540929      0
#> 170   1.056520894      0
#> 171   0.564586266      0
#> 172   0.227221739      0
#> 173  -0.751881130      0
#> 174  -1.167668197      0
#> 175   0.199289612      0
#> 176   1.910907423      0
#> 177   0.667316694      0
#> 178   0.954196270      0
#> 179   0.536569722      0
#> 180  -0.088665983      0
#> 181  -0.412917107      0
#> 182  -0.932221493      0
#> 183   1.297750884      0
#> 184  -0.375391140      0
#> 185  -1.455336608      0
#> 186   0.926505703      0
#> 187   0.144130635      0
#> 188   0.083453517      0
#> 189   0.282205443      0
#> 190   0.768135790      0
#> 191  -1.540478044      0
#> 192   0.292273223      0
#> 193   0.375738723      0
#> 194   0.403880432      0
#> 195  -0.849886506      0
#> 196   2.203330342      0
#> 197  -1.497824632      0
#> 198   0.780712508      0
#> 199  -0.559585198      0
#> 200   1.048918668      0
#> 201   1.473598596      0
#> 202   1.679762417      0
#> 203   1.028958038      0
#> 204  -1.007713548      0
#> 205  -1.405372427      0
#> 206  -0.324646663      0
#> 207  -0.223897287      0
#> 208  -0.503081221      0
#> 209   0.147005120      0
#> 210   1.285371988      0
#> 211   0.413084415      0
#> 212  -0.806225589      0
#> 213   1.547082391      0
#> 214  -0.009406531      0
#> 215   0.031007021      0
#> 216  -0.042924848      0
#> 217  -0.739832550      0
#> 218  -0.556548058      0
#> 219   2.114770737      0
#> 220  -0.785098522      0
#> 221  -0.405743654      0
#> 222  -0.311145475      0
#> 223   2.052412443      0
#> 224  -0.101236866      0
#> 225  -0.759814595      0
#> 226   0.453025289      0
#> 227  -1.632858106      0
#> 228   0.374107634      0
#> 229  -0.317903779      0
#> 230   0.010603206      0
#> 231   0.273869847      0
#> 232   0.701322772      0
#> 233   0.692609996      0
#> 234  -0.709783000      0
#> 235  -0.068727291      0
#> 236   0.711868587      0
#> 237   1.062297615      0
#> 238  -1.263156054      0
#> 239   0.126338912      0
#> 240   0.609699621      0
#> 241  -0.413235239      0
#> 242   1.015516272      0
#> 243   1.780509564      0
#> 244   2.093304573      0
#> 245  -0.635101111      0
#> 246  -0.860429846      0
#> 247   0.260247898      0
#> 248   0.157406514      0
#> 249  -1.133620582      0
#> 250   0.398171016      0
#> 251   0.397611716      0
#> 252   1.145769049      0
#> 253   0.395146375      0
#> 254  -1.507784712      0
#> 255   1.042354209      0
#> 256  -0.393374572      0
#> 257   0.681089179      0
#> 258  -1.693891981      0
#> 259  -0.553264727      0
#> 260  -1.045770580      0
#> 261  -1.624273349      0
#> 262   1.335797165      0
#> 263   2.648007054      0
#> 264  -0.240081027      0
#> 265  -0.311444679      0
#> 266   1.190432975      0
#> 267  -0.033093376      0
#> 268   0.619257656      0
#> 269  -1.480149627      0
#> 270  -0.989044191      0
#> 271  -0.273788975      0
#> 272   0.537607657      0
#> 273   0.346787457      0
#> 274  -0.425782056      0
#> 275   0.319889348      0
#> 276  -1.265089943      0
#> 277   1.479187730      0
#> 278   2.104443101      0
#> 279  -2.382577248      0
#> 280  -1.678710345      0
#> 281   0.225534359      0
#> 282  -0.053976975      0
#> 283   0.272311547      0
#> 284   0.727455926      0
#> 285  -0.541748901      0
#> 286   0.857315205      0
#> 287   0.179147148      0
#> 288  -0.462092831      0
#> 289  -0.835945615      0
#> 290  -0.662132547      0
#> 291  -0.267239446      0
#> 292  -0.513760724      0
#> 293   0.961329795      0
#> 294   1.237722644      0
#> 295   0.058006388      0
#> 296   0.437930854      0
#> 297   0.308885416      0
#> 298  -0.014618166      0
#> 299   1.701131130      0
#> 300   0.964462983      0
#> 301   0.625026020      0
#> 302   0.885637004      0
#> 303   0.456676188      0
#> 304   0.506005150      0
#> 305  -0.705628810      0
#> 306   0.601926919      0
#> 307  -0.329815239      0
#> 308   0.386045122      0
#> 309  -0.303996435      0
#> 310   2.403011088      0
#> 311  -0.695840185      0
#> 312   1.428566689      0
#> 313   0.469333648      0
#> 314  -1.903794854      0
#> 315  -1.251579700      0
#> 316   0.139088087      0
#> 317  -0.106028101      0
#> 318  -0.807757023      0
#> 319   1.577721045      0
#> 320  -0.688923385      0
#> 321   0.274456853      0
#> 322  -2.010384388      0
#> 323   0.323100983      0
#> 324  -1.566777800      0
#> 325   0.280020545      0
#> 326  -0.138252854      0
#> 327   0.270202148      0
#> 328  -0.787674324      0
#> 329  -0.950976894      0
#> 330  -0.886940240      0
#> 331  -1.698785596      0
#> 332  -0.585191897      0
#> 333   0.438111125      0
#> 334  -0.827340140      0
#> 335   0.572323290      0
#> 336  -0.028045212      0
#> 337   0.471185802      0
#> 338   1.250809494      0
#> 339   0.203031385      0
#> 340  -0.214892494      0
#> 341   2.378810931      0
#> 342   1.497964659      0
#> 343  -0.622332710      0
#> 344  -0.472798099      0
#> 345   1.727534093      0
#> 346  -0.558494566      0
#> 347  -0.167338146      0
#> 348   0.045046748      0
#> 349  -1.146752104      0
#> 350  -1.787195609      0
#> 351   0.852251153      0
#> 352   1.024445098      0
#> 353  -0.246896660      0
#> 354   2.062199117      0
#> 355  -1.930032488      0
#> 356   0.760635497      0
#> 357   1.444241895      0
#> 358  -3.638299550      0
#> 359   1.756083661      0
#> 360   0.214520920      0
#> 361   1.357570902      0
#> 362  -0.775951559      0
#> 363   0.444459659      0
#> 364  -1.158572731      0
#> 365   0.761761719      0
#> 366  -0.735430793      0
#> 367  -1.273616228      0
#> 368  -1.059591468      0
#> 369  -0.283824474      0
#> 370  -0.724562820      0
#> 371   0.109940467      0
#> 372   0.787759852      0
#> 373   0.339080074      0
#> 374   0.148932551      0
#> 375  -0.308549965      0
#> 376  -0.471503182      0
#> 377  -1.414214964      0
#> 378   0.134291454      0
#> 379  -0.288653689      0
#> 380   1.077081088      0
#> 381  -1.021652924      0
#> 382   1.835582439      0
#> 383   0.504059307      0
#> 384   0.496862815      0
#> 385  -0.772625703      0
#> 386   0.329040289      0
#> 387  -2.771827555      0
#> 388  -0.822388517      0
#> 389  -1.653020540      0
#> 390   0.384422440      0
#> 391  -2.434236290      0
#> 392  -0.008522141      0
#> 393   1.650778770      0
#> 394   1.586785711      0
#> 395   0.768321188      0
#> 396  -0.919567550      0
#> 397   0.104287446      0
#> 398   0.395177446      0
#> 399   1.098654703      0
#> 400   0.576501413      0
#> 401   0.672538436      0
#> 402   1.175537230      0
#> 403   0.564210560      0
#> 404  -0.083364550      0
#> 405  -0.015421245      0
#> 406  -0.144937979      0
#> 407  -1.433943692      0
#> 408  -1.055673291      0
#> 409   1.657298071      0
#> 410  -1.759966636      0
#> 411   1.674388420      0
#> 412  -0.643468226      0
#> 413  -0.327368915      0
#> 414  -0.625654566      0
#> 415   0.123753872      0
#> 416  -0.340782085      0
#> 417   0.109183078      0
#> 418   0.178542795      0
#> 419  -0.589623187      0
#> 420  -0.838328018      0
#> 421   1.250669348      0
#> 422  -1.273412620      0
#> 423  -0.530200435      0
#> 424   0.228126975      0
#> 425   0.247116279      0
#> 426   1.685749862      0
#> 427  -0.357063567      0
#> 428   0.231804943      0
#> 429  -0.690424660      0
#> 430   0.936902471      0
#> 431   0.589678983      0
#> 432  -0.185830744      0
#> 433  -1.379731442      0
#> 434  -2.088377549      0
#> 435   1.003612729      0
#> 436   1.085963433      0
#> 437   0.344257645      0
#> 438  -0.469464632      0
#> 439   1.785335492      0
#> 440  -1.075808833      0
#> 441   0.351667511      0
#> 442  -0.139360904      0
#> 443  -0.247477705      0
#> 444  -0.079201903      0
#> 445   0.035138044      0
#> 446   0.254971784      0
#> 447  -1.792157226      0
#> 448  -0.875451936      0
#> 449   0.433339443      0
#> 450   0.420558421      0
#> 451   0.307379836      0
#> 452  -0.842009197      0
#> 453   0.482135345      0
#> 454   0.473891626      0
#> 455   0.791100583      0
#> 456   0.252325382      0
#> 457  -0.584142653      0
#> 458  -0.457928921      0
#> 459  -0.110053213      0
#> 460   0.404902315      0
#> 461   1.253571404      0
#> 462  -0.534553407      0
#> 463   1.267470707      0
#> 464   1.993432420      0
#> 465  -1.109859083      0
#> 466   1.011173885      0
#> 467  -0.891303589      0
#> 468   0.700184687      0
#> 469  -1.531639612      0
#> 470  -0.312161490      0
#> 471   1.992509291      0
#> 472  -0.067286464      0
#> 473   0.771639479      0
#> 474   0.061986117      0
#> 475  -1.322610970      0
#> 476   0.563590863      0
#> 477  -0.878436454      0
#> 478   0.459946722      0
#> 479  -0.036393518      0
#> 480  -0.156328750      0
#> 481   2.611675400      0
#> 482   0.477080073      0
#> 483   1.469980859      0
#> 484   0.982799095      0
#> 485  -0.186425599      0
#> 486   1.455475004      0
#> 487  -0.994174628      0
#> 488   0.764535427      0
#> 489   0.635073063      0
#> 490  -1.321038967      0
#> 491   0.587266481      0
#> 492   0.364869121      0
#> 493   0.662602533      0
#> 494   0.571028300      0
#> 495  -0.481375610      0
#> 496   0.853466674      0
#> 497   0.296592422      0
#> 498   0.039385518      0
#> 499  -0.847295191      0
#> 500   0.790571164      0
#> 501  -1.015459089      0
#> 502   0.259801770      0
#> 503  -0.322637489      0
#> 504  -1.388413537      0
#> 505  -0.685230324      0
#> 506  -0.173358840      0
#> 507   0.660507183      0
#> 508  -1.074590564      0
#> 509  -0.815192278      0
#> 510  -0.316802343      0
#> 511  -0.205535176      0
#> 512   1.456643153      0
#> 513   0.656813402      0
#> 514   1.215925790      0
#> 515  -0.596081103      0
#> 516   0.166971038      0
#> 517  -0.532066684      0
#> 518  -0.102542604      0
#> 519  -1.704868182      0
#> 520   0.627633270      0
#> 521   0.923212394      0
#> 522   1.389116533      0
#> 523  -0.513992453      0
#> 524   0.178811866      0
#> 525  -0.704690848      0
#> 526  -0.883151304      0
#> 527  -0.378511464      0
#> 528  -1.234411585      0
#> 529   1.543891389      0
#> 530   0.014320784      0
#> 531   0.182416611      0
#> 532  -0.791154731      0
#> 533   1.433249858      0
#> 534   2.628680523      0
#> 535  -0.839929581      0
#> 536   0.230079224      0
#> 537   1.507706675      0
#> 538   0.430289490      0
#> 539  -2.121853979      0
#> 540  -0.356084781      0
#> 541  -1.336006601      0
#> 542  -0.891946427      0
#> 543  -1.578913083      0
#> 544   0.245807763      0
#> 545   0.285982257      0
#> 546   1.274918645      0
#> 547  -0.907398424      0
#> 548  -0.826553574      0
#> 549   0.059626180      0
#> 550   2.057173490      0
#> 551  -0.148742920      0
#> 552   0.034376984      0
#> 553   0.124664571      0
#> 554   0.505801799      0
#> 555  -0.861428385      0
#> 556  -1.692565602      0
#> 557   0.534975920      0
#> 558   0.828409727      0
#> 559   0.407844629      0
#> 560  -0.255803713      0
#> 561   0.851840931      0
#> 562   0.303357285      0
#> 563   0.086588513      0
#> 564  -0.988224324      0
#> 565  -1.601922990      0
#> 566  -1.035635791      0
#> 567  -0.020944668      0
#> 568   1.444477551      0
#> 569   0.967722351      0
#> 570  -0.478481082      0
#> 571   1.258163353      0
#> 572   0.649484908      0
#> 573  -1.598429106      0
#> 574   0.571525638      0
#> 575  -0.910940612      0
#> 576   1.113038065      0
#> 577   0.522347354      0
#> 578   0.450966017      0
#> 579   0.273987074      0
#> 580  -0.829103001      0
#> 581  -1.261721566      0
#> 582  -0.812516239      0
#> 583   0.942317199      0
#> 584   2.842977965      0
#> 585   0.572579329      0
#> 586  -0.429111295      0
#> 587   0.173837127      0
#> 588   0.737832821      0
#> 589  -1.644137897      0
#> 590   1.295594516      0
#> 591   0.680989795      0
#> 592  -1.089203184      0
#> 593  -0.965771987      0
#> 594   0.200222776      0
#> 595  -0.670606995      0
#> 596   0.348800756      0
#> 597  -0.239520450      0
#> 598  -0.134121170      0
#> 599   1.834193643      0
#> 600  -0.923136540      0
#> 601   0.052515406      0
#> 602  -1.511802650      0
#> 603   0.335273081      0
#> 604  -1.730464435      0
#> 605   1.113940511      0
#> 606   0.052461998      0
#> 607  -0.862229904      0
#> 608   0.492927759      0
#> 609  -0.028332859      0
#> 610  -0.703889595      0
#> 611   0.389413084      0
#> 612   1.193522887      0
#> 613   0.133305940      0
#> 614  -0.036313392      0
#> 615   0.134828654      0
#> 616  -0.611501172      0
#> 617  -0.791176142      0
#> 618  -0.493920459      0
#> 619   0.198643008      0
#> 620  -0.658164898      0
#> 621   0.729010249      0
#> 622  -0.156114725      0
#> 623  -0.496124831      0
#> 624   0.653731988      0
#> 625  -0.245294565      0
#> 626   0.171394224      0
#> 627  -0.430296713      0
#> 628  -0.278155573      0
#> 629  -0.296046561      0
#> 630   0.398493848      0
#> 631   1.421431916      0
#> 632   0.664748440      0
#> 633  -0.252340529      0
#> 634  -1.671757812      0
#> 635  -0.819614351      0
#> 636   1.965270457      0
#> 637   0.394686652      0
#> 638   1.907251871      0
#> 639  -0.805580001      0
#> 640   0.858004406      0
#> 641  -0.808517687      0
#> 642  -0.422742259      0
#> 643  -0.800414137      0
#> 644   1.569870367      0
#> 645   1.550702855      0
#> 646   0.415964009      0
#> 647  -0.036793188      0
#> 648   0.119440345      0
#> 649   1.330510731      0
#> 650   1.467567962      0
#> 651   1.020429981      0
#> 652  -1.196844446      0
#> 653  -0.804440685      0
#> 654  -1.142105558      0
#> 655   0.572992633      0
#> 656  -2.191924819      0
#> 657   0.588796188      0
#> 658   0.111655752      0
#> 659   1.265822482      0
#> 660  -0.641098479      0
#> 661   1.625884117      0
#> 662  -1.292909166      0
#> 663  -0.262419324      0
#> 664   0.474880479      0
#> 665   0.645795091      0
#> 666  -0.076978279      0
#> 667   0.291318878      0
#> 668  -1.080558739      0
#> 669   0.791076179      0
#> 670   0.696252009      0
#> 671  -0.194966659      0
#> 672   2.270876590      0
#> 673  -0.382918769      0
#> 674   1.367873925      0
#> 675  -0.189244208      0
#> 676  -0.081048200      0
#> 677  -0.003120947      0
#> 678   0.332517065      0
#> 679  -1.050939916      0
#> 680  -0.262056216      0
#> 681   0.970688781      0
#> 682  -1.132367756      0
#> 683   0.111255714      0
#> 684   0.852291810      0
#> 685  -1.165635351      0
#> 686   1.171964645      0
#> 687  -0.711079694      0
#> 688   0.237274693      0
#> 689   0.294669288      0
#> 690   0.339744296      0
#> 691  -0.025475770      0
#> 692  -0.993872298      0
#> 693  -0.428677836      0
#> 694  -1.217211761      0
#> 695   0.558252358      0
#> 696   1.329822326      0
#> 697  -1.241421063      0
#> 698  -0.711665220      0
#> 699  -0.828664720      0
#> 700   1.485420065      0
#> 701  -0.249313133      0
#> 702  -0.998514915      0
#> 703  -1.410382071      0
#> 704  -0.894174081      0
#> 705   0.390904504      0
#> 706   0.563446203      0
#> 707  -1.106438111      0
#> 708  -0.644253533      0
#> 709   0.125990491      0
#> 710  -1.623490561      0
#> 711  -0.206487239      0
#> 712  -0.298132722      0
#> 713   0.317720404      0
#> 714  -2.726884295      0
#> 715   0.470904123      0
#> 716   0.286241351      0
#> 717  -0.743896308      0
#> 718   0.820228539      0
#> 719   0.835693007      0
#> 720   0.709240635      0
#> 721  -0.405075155      0
#> 722   0.114580268      0
#> 723   0.588460583      0
#> 724  -1.099889980      0
#> 725   0.999090244      0
#> 726  -0.260078833      0
#> 727  -0.293757993      0
#> 728   0.033922963      0
#> 729   0.122355916      0
#> 730   0.015369254      0
#> 731  -0.303384576      0
#> 732   0.758760042      0
#> 733   0.361696479      0
#> 734  -0.353402561      0
#> 735   0.919312260      0
#> 736   1.202345268      0
#> 737   1.695528822      0
#> 738   1.307907596      0
#> 739   0.663321342      0
#> 740   0.270624521      0
#> 741   0.692984491      0
#> 742   0.473489048      0
#> 743   1.134162556      0
#> 744  -1.167888620      0
#> 745  -0.291536900      0
#> 746  -1.382475714      0
#> 747  -0.218298833      0
#> 748  -0.071965149      0
#> 749  -0.371126790      0
#> 750   2.293962399      0
#> 751   1.126584608      0
#> 752  -0.271730495      0
#> 753  -0.550824454      0
#> 754  -0.148207879      0
#> 755   0.568978534      0
#> 756  -0.090135887      0
#> 757  -0.939822885      0
#> 758   0.726837711      0
#> 759  -0.591738397      0
#> 760  -0.619542361      0
#> 761  -0.337570480      0
#> 762   0.491333728      0
#> 763   0.706373468      0
#> 764   0.548749693      0
#> 765   0.915827680      0
#> 766  -1.814185365      0
#> 767  -0.294285137      0
#> 768  -0.315010595      0
#> 769   0.385543247      0
#> 770   1.167818069      0
#> 771  -1.870903057      0
#> 772  -0.176312152      0
#> 773   1.377400656      0
#> 774  -0.187226869      0
#> 775  -0.901239564      0
#> 776  -0.802184732      0
#> 777  -1.131943996      0
#> 778  -2.010540552      0
#> 779   0.951173786      0
#> 780  -0.714158498      0
#> 781   0.441692132      0
#> 782   0.956951151      0
#> 783   1.007253037      0
#> 784   0.094770441      0
#> 785   1.226050779      0
#> 786   0.296851408      0
#> 787  -0.397394288      0
#> 788  -0.095926548      0
#> 789  -0.684692476      0
#> 790   1.657096674      0
#> 791  -0.919032855      0
#> 792  -1.177416940      0
#> 793  -0.432040892      0
#> 794   1.342314009      0
#> 795  -1.063447410      0
#> 796  -1.359386097      0
#> 797  -1.114634226      0
#> 798   0.779780251      0
#> 799   1.151645922      0
#> 800  -0.314111810      0
#> 801  -0.891916242      0
#> 802  -0.805901743      0
#> 803  -1.180632706      0
#> 804  -0.845453364      0
#> 805  -2.233304700      0
#> 806   0.637528045      0
#> 807   0.379631057      0
#> 808  -0.175857498      0
#> 809   0.335750902      0
#> 810  -0.195531479      0
#> 811   0.788765698      0
#> 812  -3.259732036      0
#> 813  -0.626598058      0
#> 814   1.358477170      0
#> 815  -0.249536551      0
#> 816  -1.206395446      0
#> 817  -1.131658891      0
#> 818   2.441791030      0
#> 819  -0.555025712      0
#> 820   0.070958770      0
#> 821  -1.073439642      0
#> 822  -0.679919857      0
#> 823   1.217286896      0
#> 824   0.019448329      0
#> 825  -0.655440567      0
#> 826   0.756043300      0
#> 827   0.181398579      0
#> 828   1.185954313      0
#> 829  -0.855653609      0
#> 830   0.795980776      0
#> 831   0.642214770      0
#> 832   0.831243360      0
#> 833  -0.169946942      0
#> 834  -0.335985095      0
#> 835  -0.604395400      0
#> 836  -0.211478048      0
#> 837  -0.249273164      0
#> 838  -0.484087724      0
#> 839   0.348645116      0
#> 840   0.332343729      0
#> 841   2.154964110      0
#> 842   1.239626819      0
#> 843   0.767066642      0
#> 844   0.471717293      0
#> 845   0.213964272      0
#> 846   1.553366606      0
#> 847  -1.097016154      0
#> 848  -1.769997257      0
#> 849   1.903216766      0
#> 850   0.594935725      0
#> 851   1.140243729      0
#> 852  -1.202717162      0
#> 853   1.984490219      0
#> 854  -0.038652560      0
#> 855   1.473644307      0
#> 856   0.568450097      0
#> 857  -0.506357889      0
#> 858   0.028173927      0
#> 859  -1.215062267      0
#> 860  -0.017918955      0
#> 861   1.713382925      0
#> 862   0.014066302      0
#> 863   1.282037200      0
#> 864  -1.300836435      0
#> 865   0.512151597      0
#> 866  -0.226019899      0
#> 867   0.914285939      0
#> 868  -1.334068884      0
#> 869   0.907337707      0
#> 870   0.052449021      0
#> 871  -0.223224447      0
#> 872   1.574910817      0
#> 873  -0.034599293      0
#> 874   0.837995214      0
#> 875  -0.553130648      0
#> 876   0.016472022      0
#> 877  -0.624856311      0
#> 878  -0.175408550      0
#> 879   0.272328166      0
#> 880   0.986421643      0
#> 881   0.360388431      0
#> 882  -0.592452995      0
#> 883   0.048456492      0
#> 884   0.259628011      0
#> 885  -1.223028930      0
#> 886  -0.703311782      0
#> 887   0.769830164      0
#> 888   0.425962523      0
#> 889  -0.184172473      0
#> 890   0.174604464      0
#> 891  -2.560457184      0
#> 892   1.180996802      0
#> 893  -0.539709400      0
#> 894  -0.434979835      0
#> 895  -0.172543487      0
#> 896  -0.075767434      0
#> 897   0.024646846      0
#> 898   0.916693774      1
#> 899   0.107256795      1
#> 900   0.973867455      1
#> 901   0.093443549      1
#> 902   1.379616773      1
#> 903  -0.031550766      1
#> 904   0.550510990      1
#> 905   1.365311208      1
#> 906   0.955937358      1
#> 907  -0.113469479      1
#> 908  -0.356325598      1
#> 909   1.405492080      1
#> 910   1.616890077      1
#> 911   0.870252295      1
#> 912   1.666521558      1
#> 913   0.811508393      1
#> 914   1.456466979      1
#> 915   1.378020103      1
#> 916  -0.809009720      1
#> 917   2.013550042      1
#> 918   0.895882123      1
#> 919   1.234771621      1
#> 920   0.380404662      1
#> 921   1.412786255      1
#> 922  -0.500985428      1
#> 923   1.927554746      1
#> 924   2.046072218      1
#> 925   0.775050709      1
#> 926   0.408823538      1
#> 927   0.171878277      1
#> 928   1.056306643      1
#> 929   1.869071907      1
#> 930   0.754769259      1
#> 931   0.332194682      1
#> 932   0.703296843      1
#> 933   1.694599447      1
#> 934   0.561374299      1
#> 935   1.398418625      1
#> 936  -0.167444337      1
#> 937  -1.133550392      1
#> 938   1.922409729      1
#> 939   0.750337252      1
#> 940   0.242226908      1
#> 941   1.578062413      1
#> 942   1.255893691      1
#> 943  -0.395690059      1
#> 944   1.713710374      1
#> 945   1.257484863      1
#> 946   0.481084829      1
#> 947   0.909848634      1
#> 948   1.875510748      1
#> 949   1.506234340      1
#> 950   1.386971607      1
#> 951   0.829932851      1
#> 952   1.397170132      1
#> 953   0.418055683      1
#> 954   0.942499189      1
#> 955   0.813163919      1
#> 956   1.279821724      1
#> 957   1.739776135      1
#> 958   0.221314395      1
#> 959   3.030814961      1
#> 960   2.096139825      1
#> 961   1.890472138      1
#> 962   0.810084756      1
#> 963   0.795823724      1
#> 964   0.634597643      1
#> 965  -0.034457862      1
#> 966   0.305565326      1
#> 967  -0.164146007      1
#> 968   1.512914312      1
#> 969  -0.017622589      1
#> 970   0.278572912      1
#> 971  -0.034416817      1
#> 972  -0.266766092      1
#> 973   1.446330972      1
#> 974   0.221858499      1
#> 975   0.797018526      1
#> 976  -0.207234093      1
#> 977   1.143496228      1
#> 978   2.258993660      1
#> 979   0.682824062      1
#> 980   0.987425782      1
#> 981   0.432846033      1
#> 982   0.619809455      1
#> 983   1.853766802      1
#> 984   1.668109802      1
#> 985   1.084024441      1
#> 986   1.382682352      1
#> 987  -0.508184216      1
#> 988  -0.424966324      1
#> 989   1.150859043      1
#> 990   0.534125905      1
#> 991   1.509236329      1
#> 992   0.632167324      1
#> 993   1.516506343      1
#> 994   0.849377123      1
#> 995   0.949335312      1
#> 996   0.994755208      1
#> 997   1.307454527      1
#> 998   0.884198256      1
#> 999   1.267443180      1
#> 1000  1.036126132      1