Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.011777550      0
#> 2     1.717762850      0
#> 3    -1.039161626      0
#> 4    -0.453535717      0
#> 5    -1.326013239      0
#> 6     0.752307363      0
#> 7     0.940452016      0
#> 8    -0.683822691      0
#> 9    -0.193103671      0
#> 10    0.523965293      0
#> 11   -1.001825794      0
#> 12    0.868884620      0
#> 13   -1.758271688      0
#> 14    0.493820311      0
#> 15   -1.308170570      0
#> 16    0.450119131      0
#> 17    0.665744859      0
#> 18   -0.198134968      0
#> 19   -0.014528873      0
#> 20    0.175005975      0
#> 21   -0.674720033      0
#> 22    1.202116385      0
#> 23   -0.167921645      0
#> 24   -0.349484245      0
#> 25   -0.209936672      0
#> 26    0.823887619      0
#> 27    0.659435437      0
#> 28    0.334030970      0
#> 29    0.409881822      0
#> 30   -0.092377582      0
#> 31    0.362384116      0
#> 32   -2.419026501      0
#> 33   -0.908377381      0
#> 34   -1.533270808      0
#> 35    0.914713306      0
#> 36   -0.585094202      0
#> 37    1.353611985      0
#> 38    0.023826048      0
#> 39   -0.609640796      0
#> 40   -0.741936951      0
#> 41   -0.339030802      0
#> 42    0.235538045      0
#> 43   -0.333268653      0
#> 44   -0.433286070      0
#> 45    0.525472031      0
#> 46    3.026616594      0
#> 47   -1.262216755      0
#> 48   -1.339397005      0
#> 49    1.070598023      0
#> 50    0.791173676      0
#> 51   -2.914962401      0
#> 52    1.284129277      0
#> 53   -0.287922913      0
#> 54   -0.257659370      0
#> 55    1.208705425      0
#> 56    0.085351729      0
#> 57    0.356951881      0
#> 58    2.548195730      0
#> 59   -0.476660683      0
#> 60   -0.360435576      0
#> 61   -0.423945078      0
#> 62    0.613367318      0
#> 63    0.202796873      0
#> 64   -1.843735892      0
#> 65   -2.094747103      0
#> 66   -0.375983803      0
#> 67    0.808480892      0
#> 68    0.311941086      0
#> 69   -0.806121063      0
#> 70    0.902171296      0
#> 71    1.758394613      0
#> 72    0.998234893      0
#> 73    0.965880488      0
#> 74   -0.125348526      0
#> 75    0.356056554      0
#> 76    0.944746330      0
#> 77    0.154519101      0
#> 78    0.268503222      0
#> 79    0.586500865      0
#> 80   -1.177378639      0
#> 81    2.489085298      0
#> 82   -0.938611188      0
#> 83   -1.370528317      0
#> 84    0.262707625      0
#> 85   -2.401385232      0
#> 86   -0.567167022      0
#> 87    0.663781398      0
#> 88   -0.967747329      0
#> 89   -0.608055638      0
#> 90   -0.349035912      0
#> 91   -0.213993220      0
#> 92   -2.315212760      0
#> 93   -2.240206987      0
#> 94    1.103400040      0
#> 95   -1.854997959      0
#> 96    0.463672737      0
#> 97    0.426555365      0
#> 98    0.240862284      0
#> 99    0.439023328      0
#> 100  -0.019461424      0
#> 101   0.935820213      0
#> 102  -0.412944937      0
#> 103   1.983100778      0
#> 104   0.333771585      0
#> 105  -0.496043379      0
#> 106  -0.407813447      0
#> 107   0.225697274      0
#> 108  -2.225328803      0
#> 109   1.003680001      0
#> 110  -1.080784244      0
#> 111   0.091542317      0
#> 112  -1.435107486      0
#> 113   0.482338775      0
#> 114  -0.677212309      0
#> 115  -0.476298831      0
#> 116  -1.506332735      0
#> 117  -1.107308700      0
#> 118   0.522505583      0
#> 119   1.019925340      0
#> 120  -1.440640541      0
#> 121   1.667824422      0
#> 122   2.231456360      0
#> 123  -0.629021371      0
#> 124   0.888692395      0
#> 125  -2.636618006      0
#> 126   1.857347200      0
#> 127   0.389414502      0
#> 128   1.934349110      0
#> 129   0.653686876      0
#> 130  -0.169157053      0
#> 131   0.909325365      0
#> 132  -0.794327279      0
#> 133  -0.093166135      0
#> 134  -0.176073785      0
#> 135  -1.600597472      0
#> 136  -0.092499487      0
#> 137  -0.034975934      0
#> 138  -0.864604631      0
#> 139  -0.887786802      0
#> 140   1.823995487      0
#> 141  -0.657010532      0
#> 142  -0.019391187      0
#> 143  -0.980545216      0
#> 144   0.664030045      0
#> 145   1.606575566      0
#> 146   0.528999887      0
#> 147  -0.749560009      0
#> 148  -0.393920946      0
#> 149  -0.087315602      0
#> 150   0.097127662      0
#> 151   1.222544456      0
#> 152  -0.434431751      0
#> 153  -0.388122911      0
#> 154   0.044043308      0
#> 155  -1.037536602      0
#> 156  -0.249995641      0
#> 157  -0.423260344      0
#> 158  -0.292403459      0
#> 159  -0.254004845      0
#> 160   1.282174238      0
#> 161   1.265095043      0
#> 162  -0.601967305      0
#> 163   1.227265046      0
#> 164   1.195843065      0
#> 165  -1.231565300      0
#> 166  -0.629680765      0
#> 167   0.904327554      0
#> 168   0.518147536      0
#> 169  -0.521461615      0
#> 170  -0.233514249      0
#> 171  -0.233830914      0
#> 172  -1.385248877      0
#> 173  -0.847259699      0
#> 174  -0.679488401      0
#> 175   1.542475310      0
#> 176  -2.323898086      0
#> 177  -0.679923797      0
#> 178  -1.692109379      0
#> 179   0.241203753      0
#> 180   1.251793242      0
#> 181  -0.450070183      0
#> 182  -1.000847903      0
#> 183   0.527267075      0
#> 184   0.423854770      0
#> 185  -0.730970395      0
#> 186  -0.490304831      0
#> 187   1.174386210      0
#> 188  -0.174834550      0
#> 189   0.380119747      0
#> 190   1.589161404      0
#> 191   1.951772386      0
#> 192   0.927269807      0
#> 193  -0.851967707      0
#> 194   0.596744930      0
#> 195  -0.777905845      0
#> 196  -0.624003935      0
#> 197  -0.409948379      0
#> 198   0.421295298      0
#> 199  -0.609875133      0
#> 200   0.786425422      0
#> 201  -0.335093925      0
#> 202   0.793124818      0
#> 203   0.633291261      0
#> 204  -1.531287197      0
#> 205   1.643331203      0
#> 206   0.483094613      0
#> 207  -0.532667374      0
#> 208   0.045009888      0
#> 209  -0.212704099      0
#> 210   0.448720035      0
#> 211  -0.529686639      0
#> 212   0.361157614      0
#> 213   0.669565902      0
#> 214   0.584332024      0
#> 215  -0.308055792      0
#> 216  -0.369451319      0
#> 217   0.935027476      0
#> 218   0.797751382      0
#> 219  -0.724955683      0
#> 220   0.571598056      0
#> 221  -1.313215255      0
#> 222  -1.459686156      0
#> 223  -0.145347572      0
#> 224  -0.500061128      0
#> 225   1.316181955      0
#> 226   0.486055189      0
#> 227  -0.977507058      0
#> 228   0.411269315      0
#> 229  -0.854300172      0
#> 230   0.586322160      0
#> 231  -0.702612836      0
#> 232  -0.844858628      0
#> 233   1.514643710      0
#> 234  -0.612877594      0
#> 235  -0.240671722      0
#> 236  -1.157844545      0
#> 237  -1.143427376      0
#> 238   0.195906208      0
#> 239  -0.975762769      0
#> 240  -0.271533778      0
#> 241   2.113480477      0
#> 242   1.594600230      0
#> 243   0.996980106      0
#> 244  -0.404494117      0
#> 245  -0.128791784      0
#> 246  -0.504071733      0
#> 247   1.475482912      0
#> 248   1.118712884      0
#> 249  -0.360529346      0
#> 250  -0.822840916      0
#> 251   0.366350562      0
#> 252   1.119427685      0
#> 253   0.697589112      0
#> 254  -0.512470981      0
#> 255   0.308767976      0
#> 256   0.363596697      0
#> 257   1.645628873      0
#> 258   0.478300780      0
#> 259   0.498996583      0
#> 260   0.900579818      0
#> 261  -1.051214473      0
#> 262   0.500238616      0
#> 263   0.695605109      0
#> 264   1.253209668      0
#> 265   0.671260185      0
#> 266   0.909304057      0
#> 267   0.234886206      0
#> 268   0.295307407      0
#> 269  -0.806008072      0
#> 270   0.179993579      0
#> 271   0.420904059      0
#> 272  -1.195178116      0
#> 273   0.703636633      0
#> 274   0.451494870      0
#> 275   1.725786536      0
#> 276   0.455811991      0
#> 277  -0.752975413      0
#> 278   0.196449854      0
#> 279  -0.001461631      0
#> 280   1.718421989      0
#> 281  -1.353559474      0
#> 282   0.354505826      0
#> 283  -0.478257404      0
#> 284  -0.169086297      0
#> 285   0.529240589      0
#> 286   1.580909665      0
#> 287   0.693124228      0
#> 288   0.289824579      0
#> 289   0.169336933      0
#> 290  -0.208320406      0
#> 291   0.126760762      0
#> 292   0.223377508      0
#> 293  -0.253786379      0
#> 294   0.113665624      0
#> 295   0.613741864      0
#> 296   0.779378171      0
#> 297  -2.039342570      0
#> 298   0.123369465      0
#> 299  -0.699181410      0
#> 300  -0.523287157      0
#> 301  -1.499654513      0
#> 302   1.225610411      0
#> 303  -0.665542284      0
#> 304   2.544971994      0
#> 305   0.173940528      0
#> 306   0.369796235      0
#> 307   0.592423095      0
#> 308  -1.808082154      0
#> 309   0.664802438      0
#> 310   0.104359334      0
#> 311  -1.466524936      0
#> 312  -0.669671854      0
#> 313   1.422946651      0
#> 314  -0.062801526      0
#> 315   0.343800288      0
#> 316   1.705968398      0
#> 317   1.495445877      0
#> 318  -0.499963706      0
#> 319   0.829733957      0
#> 320  -0.605913752      0
#> 321   0.153808130      0
#> 322   1.915682241      0
#> 323   0.044065423      0
#> 324   0.148768797      0
#> 325   0.277112211      0
#> 326   0.802976894      0
#> 327  -0.109107396      0
#> 328   0.171000059      0
#> 329   1.088325290      0
#> 330   0.834639440      0
#> 331  -0.699109056      0
#> 332  -1.559224916      0
#> 333  -1.169639744      0
#> 334   0.459112208      0
#> 335  -0.704418182      0
#> 336   1.689232265      0
#> 337   0.485450433      0
#> 338  -0.424822889      0
#> 339   1.257392662      0
#> 340  -0.516341001      0
#> 341   1.246246002      0
#> 342   0.174936403      0
#> 343  -1.588633735      0
#> 344   1.694907257      0
#> 345  -0.776767434      0
#> 346  -2.036608691      0
#> 347  -0.085591813      0
#> 348   1.118764721      0
#> 349   0.566602838      0
#> 350  -1.440462383      0
#> 351  -0.647227608      0
#> 352   1.798314849      0
#> 353   0.292479058      0
#> 354  -1.824861890      0
#> 355   0.735041105      0
#> 356   1.956358340      0
#> 357   0.107327883      0
#> 358   0.575913710      0
#> 359  -0.404822529      0
#> 360  -0.983678984      0
#> 361  -0.933716174      0
#> 362  -1.139490251      0
#> 363  -0.327526180      0
#> 364   0.348596916      0
#> 365   0.785314501      0
#> 366   1.231725633      0
#> 367   1.110713914      0
#> 368   0.276626469      0
#> 369   0.416273043      0
#> 370  -1.047868705      0
#> 371  -1.059758230      0
#> 372  -0.857779162      0
#> 373   1.663428811      0
#> 374  -0.700823268      0
#> 375  -0.914707028      0
#> 376   0.545264973      0
#> 377  -0.121134017      0
#> 378  -0.754691678      0
#> 379   1.203426024      0
#> 380  -0.607414396      0
#> 381  -0.696707232      0
#> 382   0.393506361      0
#> 383  -1.286107510      0
#> 384   1.192460128      0
#> 385   0.464124845      0
#> 386   1.724332550      0
#> 387   0.399493003      0
#> 388   0.575743478      0
#> 389   1.987698632      0
#> 390   1.282942016      0
#> 391   1.771673679      0
#> 392   1.713749680      0
#> 393  -2.030383701      0
#> 394   1.734780712      0
#> 395  -0.057474896      0
#> 396   0.126303429      0
#> 397  -2.500140009      0
#> 398  -0.362312563      0
#> 399   0.344404378      0
#> 400  -0.135336698      0
#> 401  -1.349073464      0
#> 402   0.863898214      0
#> 403  -0.380826975      0
#> 404  -0.618836324      0
#> 405   0.745355377      0
#> 406   1.998986439      0
#> 407   0.409142052      0
#> 408   0.879210364      0
#> 409   1.354029054      0
#> 410   0.067819410      0
#> 411   0.489351374      0
#> 412  -0.526679658      0
#> 413  -0.512087122      0
#> 414  -0.110298699      0
#> 415   0.593434051      0
#> 416   0.741298165      0
#> 417   1.185550363      0
#> 418   1.239928312      0
#> 419  -0.523475081      0
#> 420   0.032414107      0
#> 421   0.259607086      0
#> 422   0.476586909      0
#> 423  -0.016613336      0
#> 424  -1.726158491      0
#> 425  -0.365465056      0
#> 426   1.015060877      0
#> 427   0.113603541      0
#> 428   0.391050227      0
#> 429  -1.445091608      0
#> 430   1.420829130      0
#> 431  -0.523442218      0
#> 432  -1.998420007      0
#> 433  -0.330222768      0
#> 434  -0.428455493      0
#> 435   0.691199649      0
#> 436  -1.146725150      0
#> 437  -1.024923098      0
#> 438  -0.370939738      0
#> 439   0.696254231      0
#> 440   0.768854601      0
#> 441   0.527927834      0
#> 442   0.825374762      0
#> 443   1.695887554      0
#> 444   0.027003413      0
#> 445  -0.969871336      0
#> 446   0.684416797      0
#> 447  -0.836242788      0
#> 448  -1.214916568      0
#> 449  -0.599141406      0
#> 450   0.623274618      0
#> 451   1.576256014      0
#> 452   0.036850199      0
#> 453   1.382304165      0
#> 454   1.062911737      0
#> 455  -0.593528880      0
#> 456  -1.325616332      0
#> 457  -0.311621961      0
#> 458   0.890748716      0
#> 459  -0.351678457      0
#> 460   0.050829132      0
#> 461   0.328408640      0
#> 462  -0.475588975      0
#> 463  -0.362666667      0
#> 464   0.551652843      0
#> 465   0.263966797      0
#> 466  -1.171636811      0
#> 467  -0.543253176      0
#> 468   3.183838505      0
#> 469   1.448179023      0
#> 470   0.637078522      0
#> 471   0.561306014      0
#> 472   1.036217955      0
#> 473   0.851728436      0
#> 474  -0.458888503      0
#> 475  -1.395304701      0
#> 476   2.516865850      0
#> 477   0.841083861      0
#> 478   0.481728176      0
#> 479  -0.523461176      0
#> 480   2.187590767      0
#> 481   0.265260793      0
#> 482  -0.792905863      0
#> 483   0.706892751      0
#> 484  -0.600653502      0
#> 485  -0.540271458      0
#> 486   1.379041031      0
#> 487  -0.571839117      0
#> 488   0.043777363      0
#> 489  -1.382216620      0
#> 490   1.157183468      0
#> 491  -1.563844451      0
#> 492  -1.280816409      0
#> 493   1.175073760      0
#> 494   0.421903977      0
#> 495   1.289016550      0
#> 496  -1.296489966      0
#> 497   0.566912861      0
#> 498   2.107486813      0
#> 499   0.490488054      0
#> 500   0.624054377      0
#> 501   0.051253496      0
#> 502  -0.554861856      0
#> 503  -1.494789310      0
#> 504  -1.162061397      0
#> 505   0.363262607      0
#> 506  -0.797537037      0
#> 507   0.656428191      0
#> 508  -0.055225346      0
#> 509  -0.520871747      0
#> 510  -1.053241417      0
#> 511   0.372507532      0
#> 512   0.274235060      0
#> 513  -0.801075239      0
#> 514  -0.283220165      0
#> 515   0.762735980      0
#> 516  -0.399690669      0
#> 517  -0.379701441      0
#> 518  -0.228907122      0
#> 519  -0.080558067      0
#> 520  -0.717977506      0
#> 521  -0.087990270      0
#> 522  -0.456808715      0
#> 523   1.001466638      0
#> 524  -0.076818780      0
#> 525  -0.219015998      0
#> 526  -0.504990119      0
#> 527  -0.754954197      0
#> 528  -0.631038868      0
#> 529  -0.052892104      0
#> 530  -0.036009440      0
#> 531  -1.191860086      0
#> 532  -0.775450078      0
#> 533  -0.233236231      0
#> 534   2.442370175      0
#> 535  -0.194735041      0
#> 536  -0.760630783      0
#> 537  -1.441002083      0
#> 538  -0.865395108      0
#> 539  -0.421657502      0
#> 540   0.063879350      0
#> 541   1.094979989      0
#> 542  -2.130869127      0
#> 543   0.512563361      0
#> 544  -0.546171858      0
#> 545   0.054280037      0
#> 546  -2.099508892      0
#> 547   1.793449244      0
#> 548   1.185131522      0
#> 549   1.252600609      0
#> 550   0.254420223      0
#> 551  -3.054958609      0
#> 552   0.411080564      0
#> 553  -0.725623899      0
#> 554   1.030547996      0
#> 555   0.654094154      0
#> 556  -0.913243812      0
#> 557   0.130241310      0
#> 558   0.211263598      0
#> 559   0.532344192      0
#> 560  -1.418384985      0
#> 561   0.642660445      0
#> 562  -0.684744736      0
#> 563  -0.365313977      0
#> 564  -0.518165115      0
#> 565  -0.741500276      0
#> 566  -0.932052921      0
#> 567   0.454432483      0
#> 568   0.390356281      0
#> 569  -0.375987971      0
#> 570   1.643443814      0
#> 571   0.784739209      0
#> 572  -0.523871973      0
#> 573  -0.524940902      0
#> 574  -0.932840414      0
#> 575   0.714441568      0
#> 576  -0.621818651      0
#> 577  -0.927783882      0
#> 578   0.019808391      0
#> 579   0.497594505      0
#> 580   2.748120938      0
#> 581  -0.515967866      0
#> 582   0.493780012      0
#> 583  -0.212637439      0
#> 584   0.890786126      0
#> 585   1.176595018      0
#> 586   0.230081704      0
#> 587  -0.253070672      0
#> 588  -0.541533562      0
#> 589  -1.147628979      0
#> 590  -0.796569772      0
#> 591   0.528193139      0
#> 592   0.357915323      0
#> 593   0.456528344      0
#> 594   1.187592569      0
#> 595   2.466117819      0
#> 596  -0.240135072      0
#> 597  -0.257968755      0
#> 598  -1.446219211      0
#> 599   0.447936390      0
#> 600   0.215228424      0
#> 601  -1.579994078      0
#> 602  -0.499611501      0
#> 603   0.661984220      0
#> 604  -1.025834852      0
#> 605   0.656153317      0
#> 606   1.081728820      0
#> 607  -0.766124920      0
#> 608  -0.103688795      0
#> 609   0.038024741      0
#> 610   0.238719685      0
#> 611  -1.343919154      0
#> 612   0.995775319      0
#> 613   0.904240423      0
#> 614  -0.222026934      0
#> 615  -0.347745404      0
#> 616  -1.269466986      0
#> 617   0.160310264      0
#> 618  -0.209049902      0
#> 619   0.639023420      0
#> 620   0.985407935      0
#> 621   1.378276212      0
#> 622  -0.109573545      0
#> 623  -0.343166090      0
#> 624   0.377154497      0
#> 625   0.311975442      0
#> 626   0.491578980      0
#> 627   0.818701695      0
#> 628   0.252738725      0
#> 629   0.218180103      0
#> 630   0.983791186      0
#> 631  -0.672203518      0
#> 632   1.259543850      0
#> 633  -0.314355767      0
#> 634  -0.752515046      0
#> 635  -1.053216833      0
#> 636   0.434625569      0
#> 637  -1.125588346      0
#> 638  -1.735031395      0
#> 639  -0.374161228      0
#> 640  -0.458151459      0
#> 641  -1.364642048      0
#> 642  -0.082446406      0
#> 643   0.674081103      0
#> 644   0.083126655      0
#> 645   0.086880997      0
#> 646  -0.304418103      0
#> 647   0.251764428      0
#> 648  -1.032761274      0
#> 649  -0.681943801      0
#> 650   0.025856562      0
#> 651  -0.088869751      0
#> 652   0.554248421      0
#> 653  -0.197202532      0
#> 654   1.208455885      0
#> 655   1.445874930      0
#> 656   0.728099107      0
#> 657   0.762204720      0
#> 658  -0.855259308      0
#> 659   0.815331639      0
#> 660  -0.850990085      0
#> 661  -0.315471779      0
#> 662   0.536064121      0
#> 663   1.061789596      0
#> 664  -1.296448219      0
#> 665  -1.019669082      0
#> 666   1.383588657      0
#> 667   1.447809292      0
#> 668   1.007506620      0
#> 669   1.430964456      0
#> 670  -0.317624702      0
#> 671  -0.004113452      0
#> 672  -0.675699385      0
#> 673   0.175226632      0
#> 674   0.047208705      0
#> 675   0.275626267      0
#> 676   2.008165022      0
#> 677  -0.606670122      0
#> 678   0.855575231      0
#> 679  -2.205380066      0
#> 680   0.458318265      0
#> 681  -0.398485519      0
#> 682  -0.251994201      0
#> 683  -0.168559621      0
#> 684   2.697737804      0
#> 685  -0.115178256      0
#> 686  -0.108302412      0
#> 687  -0.008351913      0
#> 688  -1.294937755      0
#> 689  -0.781637525      0
#> 690   0.700732664      0
#> 691  -0.816895707      0
#> 692  -0.229872667      0
#> 693  -0.614367677      0
#> 694  -0.370796553      0
#> 695  -0.083208566      0
#> 696   0.707623993      0
#> 697   1.089874996      0
#> 698   0.350163080      0
#> 699  -0.378603000      0
#> 700   2.626730513      0
#> 701  -0.456057205      0
#> 702   0.009904795      0
#> 703  -0.197249065      0
#> 704  -0.962763195      0
#> 705  -0.996862606      0
#> 706  -0.372580796      0
#> 707   1.444858583      0
#> 708   0.206732344      0
#> 709   0.704387642      0
#> 710  -1.673209949      0
#> 711   0.305832823      0
#> 712  -1.271530822      0
#> 713   0.336672390      0
#> 714   0.069150548      0
#> 715  -1.012377441      0
#> 716  -0.315089686      0
#> 717   0.317979139      0
#> 718   1.107939580      0
#> 719   0.248103132      0
#> 720   1.544167583      0
#> 721  -0.014728742      0
#> 722   0.513864238      0
#> 723   0.338323637      0
#> 724  -1.761511842      0
#> 725  -0.043802242      0
#> 726  -1.009668566      0
#> 727  -0.087243970      0
#> 728  -0.200826446      0
#> 729   0.264884601      0
#> 730   0.419053946      0
#> 731   0.449778969      0
#> 732   1.533172217      0
#> 733  -2.005970725      0
#> 734   0.670522244      0
#> 735  -0.119819812      0
#> 736  -0.391447876      0
#> 737  -0.492374151      0
#> 738  -0.084429034      0
#> 739   0.638805422      0
#> 740   0.834371498      0
#> 741   0.431550018      0
#> 742   1.263275412      0
#> 743   0.955328209      0
#> 744   0.084481963      0
#> 745   0.359914486      0
#> 746   0.419380792      0
#> 747  -1.860276925      0
#> 748  -1.429679622      0
#> 749  -1.363472000      0
#> 750  -0.221147935      0
#> 751   0.388090137      0
#> 752   1.993976767      0
#> 753  -0.435454116      0
#> 754  -0.860938460      0
#> 755  -0.479877294      0
#> 756  -0.989532246      0
#> 757   0.717831755      0
#> 758   0.779026440      0
#> 759  -1.766969589      0
#> 760   0.219832951      0
#> 761   1.114588298      0
#> 762  -0.993620618      0
#> 763  -0.103217538      0
#> 764  -0.777362796      0
#> 765   0.224125105      0
#> 766  -0.754840029      0
#> 767  -0.635768689      0
#> 768   0.614660389      0
#> 769  -0.071368449      0
#> 770   0.901243837      0
#> 771   0.147345964      0
#> 772   1.082976140      0
#> 773   1.266195236      0
#> 774   0.771094597      0
#> 775   0.752307184      0
#> 776  -1.392122027      0
#> 777  -0.503139257      0
#> 778  -0.665840794      0
#> 779   1.801996121      0
#> 780  -0.767533362      0
#> 781   0.269296096      0
#> 782  -0.480967231      0
#> 783   0.247844077      0
#> 784   0.559555718      0
#> 785   0.155456611      0
#> 786  -0.385863528      0
#> 787  -1.380739137      0
#> 788   0.566897776      0
#> 789   0.832336629      0
#> 790   0.852204295      0
#> 791   1.156551807      0
#> 792   0.152735417      0
#> 793   1.041888751      0
#> 794   0.339066434      0
#> 795   0.856575911      0
#> 796   1.004913022      0
#> 797  -1.299435790      0
#> 798   0.928168490      0
#> 799   1.562924588      0
#> 800  -0.957252830      0
#> 801  -0.573330798      0
#> 802  -0.712965976      0
#> 803   0.321682272      0
#> 804  -0.342820101      0
#> 805  -0.830539044      0
#> 806  -2.045360027      0
#> 807   1.971281995      0
#> 808  -1.880526680      0
#> 809   1.178381883      0
#> 810   0.710188480      0
#> 811   0.471385897      0
#> 812   0.489281636      0
#> 813   0.800368762      0
#> 814   0.142782886      0
#> 815  -2.993943086      0
#> 816  -0.190548511      0
#> 817   0.674164502      0
#> 818   0.066939377      0
#> 819   1.201764221      0
#> 820   1.979897457      0
#> 821   1.313863512      0
#> 822   0.878291392      0
#> 823   1.848328179      0
#> 824  -0.449676249      0
#> 825   0.375628453      0
#> 826   0.889412196      0
#> 827   0.168091566      0
#> 828  -0.093223365      0
#> 829   0.065445165      0
#> 830   0.643587749      0
#> 831  -0.340814513      0
#> 832   1.010932914      0
#> 833   1.065221932      0
#> 834  -0.017096267      0
#> 835  -0.799840759      0
#> 836  -0.516451318      0
#> 837  -1.584319419      0
#> 838  -0.607115970      0
#> 839  -0.418513278      0
#> 840  -0.023776825      0
#> 841  -1.373659518      0
#> 842  -0.179258625      0
#> 843   1.199843164      0
#> 844  -0.159976949      0
#> 845   0.150130309      0
#> 846   0.141555361      0
#> 847   0.145283584      0
#> 848   0.361056411      0
#> 849   1.380439104      0
#> 850  -0.452610168      0
#> 851  -0.844334099      0
#> 852   1.540606340      0
#> 853  -0.862182178      0
#> 854  -0.262003259      0
#> 855  -1.501607926      0
#> 856  -1.903504265      0
#> 857   0.793208426      0
#> 858  -0.834322879      0
#> 859   0.870133090      0
#> 860   0.643036983      0
#> 861   0.300076688      0
#> 862   1.280514162      0
#> 863  -0.768933675      0
#> 864  -1.304574929      0
#> 865  -0.832316440      0
#> 866   1.541471443      0
#> 867   0.961119937      0
#> 868  -0.634460925      0
#> 869  -0.369222824      0
#> 870   1.037592275      0
#> 871  -0.038420399      0
#> 872  -0.088187573      0
#> 873   0.200583777      0
#> 874  -1.531809361      0
#> 875  -0.443945918      0
#> 876   1.041496102      0
#> 877  -0.594803621      0
#> 878  -1.370976229      0
#> 879  -0.284736541      0
#> 880   1.188599273      0
#> 881   2.050455355      0
#> 882  -0.618202074      0
#> 883  -1.353676477      0
#> 884  -0.017876334      0
#> 885   2.463790745      0
#> 886   0.340396216      0
#> 887   0.204475319      0
#> 888   2.529104670      0
#> 889   0.104576639      0
#> 890  -0.478668643      0
#> 891   1.980669986      0
#> 892  -0.446479969      0
#> 893  -0.925911512      0
#> 894  -0.969380987      0
#> 895   1.432698250      0
#> 896  -0.722003893      0
#> 897  -0.688797453      0
#> 898  -0.614507773      0
#> 899  -0.017237018      0
#> 900  -1.650402513      0
#> 901  -0.989345008      0
#> 902  -0.581584999      0
#> 903  -1.021305980      0
#> 904   1.068009759      0
#> 905  -1.364993181      0
#> 906   1.143001326      0
#> 907  -0.662363926      0
#> 908   0.695720598      0
#> 909   0.911542621      1
#> 910   1.735774874      1
#> 911   0.054075242      1
#> 912   1.306196755      1
#> 913   2.342352432      1
#> 914   1.770319817      1
#> 915   0.542803692      1
#> 916   0.180148693      1
#> 917   0.845861848      1
#> 918   0.482972081      1
#> 919   1.059417842      1
#> 920   2.089249244      1
#> 921   1.215969164      1
#> 922   1.054617078      1
#> 923   0.862036726      1
#> 924  -0.023025767      1
#> 925   2.310628721      1
#> 926   0.825362171      1
#> 927   1.505035625      1
#> 928   1.272686660      1
#> 929  -0.368309702      1
#> 930   0.107065408      1
#> 931   3.051852830      1
#> 932  -0.529033928      1
#> 933  -0.188383519      1
#> 934   1.837351526      1
#> 935   1.235338375      1
#> 936   0.693636393      1
#> 937   0.923849643      1
#> 938   1.206037935      1
#> 939   0.482433753      1
#> 940   2.465324037      1
#> 941   0.761876924      1
#> 942   0.592273355      1
#> 943   1.153154956      1
#> 944   2.104480813      1
#> 945   0.865484509      1
#> 946   0.464951846      1
#> 947   1.269800500      1
#> 948   0.285438936      1
#> 949   0.529784694      1
#> 950   0.072003062      1
#> 951  -0.500967254      1
#> 952  -0.969168121      1
#> 953   1.537896603      1
#> 954   0.754872434      1
#> 955   1.427381055      1
#> 956   1.729449770      1
#> 957  -0.562111062      1
#> 958   0.923195348      1
#> 959  -1.181122079      1
#> 960  -0.394766767      1
#> 961   1.291148965      1
#> 962   2.212022333      1
#> 963  -0.130103618      1
#> 964   0.016486531      1
#> 965   1.172644288      1
#> 966  -0.953976923      1
#> 967   1.221371226      1
#> 968   1.669947389      1
#> 969   0.974089165      1
#> 970   1.391495495      1
#> 971   0.848449638      1
#> 972   1.207828806      1
#> 973  -0.243133146      1
#> 974  -2.287933268      1
#> 975  -0.553744569      1
#> 976   0.927704215      1
#> 977   0.788927059      1
#> 978   1.760376357      1
#> 979   0.388844314      1
#> 980  -0.247351056      1
#> 981   2.551868715      1
#> 982   1.798921945      1
#> 983   0.570751189      1
#> 984  -0.730880535      1
#> 985   1.592998062      1
#> 986   0.809044041      1
#> 987   0.738886363      1
#> 988   0.513525384      1
#> 989   1.817798552      1
#> 990  -0.550781231      1
#> 991  -0.124507222      1
#> 992   0.690735616      1
#> 993   0.615978260      1
#> 994  -0.988667078      1
#> 995  -0.477020543      1
#> 996   0.077264693      1
#> 997  -1.270771638      1
#> 998   2.713967731      1
#> 999   0.651920407      1
#> 1000  1.361137993      1