Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                x actual
#> 1     1.74836888      0
#> 2     1.50467764      0
#> 3     0.72166567      0
#> 4     1.50842661      0
#> 5     0.60216244      0
#> 6     0.42244879      0
#> 7     0.83293048      0
#> 8     0.63333061      0
#> 9    -0.87896554      0
#> 10   -1.05073940      0
#> 11   -0.41485600      0
#> 12    1.31432783      0
#> 13    0.72406983      0
#> 14   -2.13778833      0
#> 15   -0.96061992      0
#> 16   -1.40269362      0
#> 17   -0.87422725      0
#> 18   -0.67050198      0
#> 19    0.98654990      0
#> 20   -0.72003666      0
#> 21   -1.98620122      0
#> 22   -1.71745172      0
#> 23   -0.60442477      0
#> 24    0.62574727      0
#> 25    1.61375811      0
#> 26    1.60548552      0
#> 27    1.04119575      0
#> 28   -0.85730259      0
#> 29    0.51550820      0
#> 30    0.61774650      0
#> 31   -0.42826979      0
#> 32    0.17877129      0
#> 33    1.68546019      0
#> 34   -1.28606595      0
#> 35   -0.62651623      0
#> 36   -0.34284559      0
#> 37   -0.39354104      0
#> 38    0.96257505      0
#> 39   -1.33840903      0
#> 40    1.28674710      0
#> 41    1.35165424      0
#> 42   -1.01479182      0
#> 43   -0.23760711      0
#> 44   -0.72621288      0
#> 45    0.17917286      0
#> 46    0.92228636      0
#> 47   -1.04653052      0
#> 48   -0.54078529      0
#> 49   -0.68058430      0
#> 50    0.15125402      0
#> 51   -0.91304180      0
#> 52    0.99512663      0
#> 53    0.53215521      0
#> 54    1.09870990      0
#> 55    0.25674663      0
#> 56    0.73652901      0
#> 57   -1.19231073      0
#> 58    0.24534623      0
#> 59    0.56372506      0
#> 60    0.69158010      0
#> 61   -0.56244536      0
#> 62   -1.30335228      0
#> 63   -0.10172910      0
#> 64    0.84634224      0
#> 65    0.81426146      0
#> 66   -0.54315138      0
#> 67    1.12697096      0
#> 68    0.44094674      0
#> 69   -1.45433016      0
#> 70    1.64497907      0
#> 71    0.37403239      0
#> 72   -0.17193342      0
#> 73   -0.50633209      0
#> 74   -0.80010055      0
#> 75    1.86665805      0
#> 76   -0.14065367      0
#> 77    0.94834999      0
#> 78   -0.77431355      0
#> 79    0.89345174      0
#> 80    1.42078560      0
#> 81   -1.37277307      0
#> 82    0.14225107      0
#> 83   -0.44736194      0
#> 84    1.64169507      0
#> 85    1.38365173      0
#> 86   -0.43532763      0
#> 87   -1.22407128      0
#> 88   -1.21179923      0
#> 89   -2.05046015      0
#> 90   -0.88768321      0
#> 91    1.15411389      0
#> 92    0.87678946      0
#> 93    0.34704202      0
#> 94   -1.02608605      0
#> 95   -1.05059199      0
#> 96    0.04690887      0
#> 97    0.26981156      0
#> 98   -0.14334953      0
#> 99   -0.94397186      0
#> 100   0.72494134      0
#> 101  -0.93180775      0
#> 102  -0.08647646      0
#> 103   0.46090613      0
#> 104   0.38446583      0
#> 105  -1.02104384      0
#> 106  -0.75423064      0
#> 107  -0.96171755      0
#> 108   0.07142550      0
#> 109   0.24614673      0
#> 110   0.82929313      0
#> 111  -0.10683436      0
#> 112  -1.53196804      0
#> 113  -0.43517341      0
#> 114  -0.14541988      0
#> 115  -0.08422141      0
#> 116  -0.25241730      0
#> 117  -1.15728405      0
#> 118  -0.90605049      0
#> 119   2.05923150      0
#> 120  -0.44423569      0
#> 121  -0.34605332      0
#> 122   0.43507492      0
#> 123  -0.44212328      0
#> 124  -0.47230145      0
#> 125  -1.50723153      0
#> 126  -0.33295795      0
#> 127  -0.40633327      0
#> 128  -0.34754772      0
#> 129  -0.60723405      0
#> 130   0.31679052      0
#> 131  -0.80923260      0
#> 132  -0.47678192      0
#> 133   1.41956698      0
#> 134   0.83029452      0
#> 135  -0.03770551      0
#> 136  -1.96991929      0
#> 137  -1.13174084      0
#> 138  -0.35215375      0
#> 139   1.34464528      0
#> 140  -0.61791392      0
#> 141   1.38463046      0
#> 142   0.38779706      0
#> 143  -0.52935884      0
#> 144  -0.04851888      0
#> 145   1.36511569      0
#> 146  -0.49141695      0
#> 147  -0.46699185      0
#> 148  -1.10626915      0
#> 149  -0.84372884      0
#> 150   0.90016226      0
#> 151   0.26906265      0
#> 152  -0.88162642      0
#> 153   0.15403503      0
#> 154  -1.37291224      0
#> 155   0.60738047      0
#> 156   0.90390996      0
#> 157   1.49723360      0
#> 158  -0.12197921      0
#> 159   0.76253787      0
#> 160   1.20326130      0
#> 161   1.90816933      0
#> 162   0.69068842      0
#> 163  -2.07216678      0
#> 164   0.63101611      0
#> 165  -1.01577160      0
#> 166  -0.89221056      0
#> 167  -0.79311865      0
#> 168   0.84755080      0
#> 169   1.30146011      0
#> 170   1.11023044      0
#> 171   0.67505361      0
#> 172   0.21771109      0
#> 173   0.53654358      0
#> 174   0.44853011      0
#> 175  -0.15203606      0
#> 176   1.01214494      0
#> 177   2.38111882      0
#> 178  -0.16183266      0
#> 179   0.63276938      0
#> 180  -0.80704296      0
#> 181  -0.11759896      0
#> 182   0.47323824      0
#> 183  -0.15370245      0
#> 184   0.35038927      0
#> 185   2.40017490      0
#> 186   0.42321434      0
#> 187   0.98332425      0
#> 188  -0.73020464      0
#> 189  -0.27697979      0
#> 190  -0.35386053      0
#> 191  -0.48252126      0
#> 192   0.43799583      0
#> 193   1.85348517      0
#> 194  -0.34208671      0
#> 195  -0.21710853      0
#> 196  -1.53218796      0
#> 197  -0.90582379      0
#> 198  -0.77324158      0
#> 199  -0.29064907      0
#> 200  -0.23132658      0
#> 201  -0.53568449      0
#> 202   1.31630873      0
#> 203  -0.70298236      0
#> 204  -1.09976307      0
#> 205   0.23293365      0
#> 206  -0.27902613      0
#> 207   1.16236310      0
#> 208  -1.18850529      0
#> 209   0.54951511      0
#> 210  -0.71594941      0
#> 211   0.59470485      0
#> 212  -0.89867934      0
#> 213  -1.14096969      0
#> 214  -0.13749374      0
#> 215   0.08752960      0
#> 216  -1.46900876      0
#> 217  -2.42633634      0
#> 218   0.64855364      0
#> 219  -1.11461939      0
#> 220  -0.74937902      0
#> 221  -0.28394600      0
#> 222   0.15636353      0
#> 223   0.78899645      0
#> 224  -0.11197189      0
#> 225  -1.76480091      0
#> 226  -1.17900636      0
#> 227  -0.53283866      0
#> 228   0.96194026      0
#> 229   1.01624165      0
#> 230   0.17126325      0
#> 231  -0.21212941      0
#> 232  -0.09830942      0
#> 233   0.67636370      0
#> 234  -1.11074595      0
#> 235  -2.08879567      0
#> 236  -0.22105850      0
#> 237   0.87402437      0
#> 238   1.31032502      0
#> 239   1.49860083      0
#> 240  -0.15312492      0
#> 241   1.29253769      0
#> 242  -0.67906742      0
#> 243   0.37756101      0
#> 244  -0.67234415      0
#> 245   1.26345358      0
#> 246   1.79301630      0
#> 247  -0.56760665      0
#> 248   0.43588468      0
#> 249   1.06119667      0
#> 250   0.71746253      0
#> 251   0.40466643      0
#> 252   2.00191233      0
#> 253   0.56596560      0
#> 254  -0.43934351      0
#> 255   0.66791726      0
#> 256  -0.13957038      0
#> 257   0.41584571      0
#> 258   1.13982053      0
#> 259  -0.54211112      0
#> 260  -0.37600903      0
#> 261   0.39246359      0
#> 262   0.63271255      0
#> 263  -0.42820157      0
#> 264   0.82670425      0
#> 265   0.66920940      0
#> 266  -1.42377196      0
#> 267  -0.08038702      0
#> 268   1.77967323      0
#> 269  -1.23876571      0
#> 270  -0.37925143      0
#> 271  -0.06783023      0
#> 272   0.26381512      0
#> 273  -1.64552140      0
#> 274  -0.73651016      0
#> 275  -0.96928293      0
#> 276  -0.67997370      0
#> 277  -1.03845053      0
#> 278   0.58432824      0
#> 279   2.24777755      0
#> 280  -1.77588873      0
#> 281   1.10274481      0
#> 282  -0.95361602      0
#> 283  -0.17960882      0
#> 284  -0.71982782      0
#> 285  -1.32957887      0
#> 286   2.10171465      0
#> 287  -0.52075657      0
#> 288   0.29272831      0
#> 289   0.36140404      0
#> 290   0.55201567      0
#> 291  -0.46479068      0
#> 292   1.33074266      0
#> 293   1.67910022      0
#> 294  -0.13185077      0
#> 295  -0.84193411      0
#> 296  -0.06018896      0
#> 297   0.76643764      0
#> 298  -1.23584985      0
#> 299   0.46530728      0
#> 300  -0.03124532      0
#> 301   0.56821823      0
#> 302   0.60408965      0
#> 303  -0.47034325      0
#> 304  -0.82728295      0
#> 305  -1.88304289      0
#> 306   0.70613517      0
#> 307   0.27032399      0
#> 308   0.54793933      0
#> 309  -0.12137369      0
#> 310  -0.74062960      0
#> 311  -0.40485345      0
#> 312   1.84905263      0
#> 313   1.85868686      0
#> 314  -0.81283855      0
#> 315  -0.15591262      0
#> 316   0.56559483      0
#> 317   0.97060511      0
#> 318   0.04925030      0
#> 319   0.25296731      0
#> 320   0.58792320      0
#> 321  -1.66623321      0
#> 322   0.32690020      0
#> 323   1.69829493      0
#> 324  -0.28722205      0
#> 325   0.24263361      0
#> 326  -0.37880810      0
#> 327  -0.28499089      0
#> 328   1.75997811      0
#> 329  -0.12762291      0
#> 330   1.12584516      0
#> 331   2.12826090      0
#> 332   2.46177387      0
#> 333  -1.00673023      0
#> 334  -2.65090672      0
#> 335  -0.71766013      0
#> 336   1.08786809      0
#> 337   0.24934598      0
#> 338   0.25371714      0
#> 339   0.95709067      0
#> 340  -1.54415527      0
#> 341  -0.86665512      0
#> 342  -0.46942217      0
#> 343   0.30851370      0
#> 344   0.46382675      0
#> 345  -0.53956684      0
#> 346  -0.46540230      0
#> 347   0.61569137      0
#> 348   1.45357974      0
#> 349   0.31540074      0
#> 350   0.29107299      0
#> 351  -0.98635881      0
#> 352   2.67490926      0
#> 353   0.63105475      0
#> 354   0.48091425      0
#> 355   0.27097170      0
#> 356   1.66872660      0
#> 357  -1.71066129      0
#> 358  -0.65674464      0
#> 359  -1.52113430      0
#> 360   0.08149353      0
#> 361   0.88635038      0
#> 362  -0.05970365      0
#> 363  -0.52401263      0
#> 364   0.38433173      0
#> 365   1.19270459      0
#> 366  -0.27483842      0
#> 367  -0.82161972      0
#> 368  -0.20561690      0
#> 369   1.59815629      0
#> 370   3.12766150      0
#> 371  -0.03396332      0
#> 372  -0.97657520      0
#> 373   0.65472118      0
#> 374   0.98925982      0
#> 375  -0.73440311      0
#> 376   0.12511109      0
#> 377  -0.87058693      0
#> 378  -1.37769417      0
#> 379   1.81344090      0
#> 380   1.20908204      0
#> 381  -0.63400006      0
#> 382   0.06207643      0
#> 383   0.32156822      0
#> 384  -1.40007259      0
#> 385   0.68525332      0
#> 386  -2.36255674      0
#> 387  -0.34949655      0
#> 388   1.62334624      0
#> 389  -0.56732962      0
#> 390  -1.46337028      0
#> 391   1.47771552      0
#> 392  -0.24989911      0
#> 393  -0.54164276      0
#> 394   0.74190281      0
#> 395   0.91216615      0
#> 396  -0.31741273      0
#> 397   1.04960528      0
#> 398  -0.70498424      0
#> 399   0.43171701      0
#> 400  -2.49175138      0
#> 401   1.32224298      0
#> 402  -1.20648764      0
#> 403   0.40770325      0
#> 404  -0.13519804      0
#> 405  -0.81494562      0
#> 406  -0.12329479      0
#> 407  -1.50054016      0
#> 408   0.47190032      0
#> 409   1.47676148      0
#> 410   0.21045756      0
#> 411   0.50905410      0
#> 412  -0.37145391      0
#> 413   0.62413015      0
#> 414  -0.30924107      0
#> 415  -1.86907484      0
#> 416  -0.29623798      0
#> 417  -0.54047424      0
#> 418   0.55536943      0
#> 419   1.24711587      0
#> 420  -0.65489683      0
#> 421   0.50211042      0
#> 422   0.80624783      0
#> 423  -0.72290487      0
#> 424  -0.71353740      0
#> 425  -0.40169636      0
#> 426   1.17639110      0
#> 427   0.91114362      0
#> 428   0.49907958      0
#> 429   0.07200161      0
#> 430   0.50490956      0
#> 431   1.48986941      0
#> 432  -1.57102962      0
#> 433  -1.59998331      0
#> 434   0.90920325      0
#> 435   0.26099769      0
#> 436   0.54402381      0
#> 437  -1.86223542      0
#> 438  -0.64659140      0
#> 439   1.13495898      0
#> 440  -0.61234408      0
#> 441   0.50306345      0
#> 442  -0.58909144      0
#> 443   0.56137593      0
#> 444   0.52878039      0
#> 445  -0.80515027      0
#> 446  -1.73012072      0
#> 447  -0.38866441      0
#> 448  -0.59780401      0
#> 449   0.70417010      0
#> 450  -0.59164187      0
#> 451  -1.22535279      0
#> 452  -0.20887078      0
#> 453   1.08615421      0
#> 454   0.76014059      0
#> 455   1.02424775      0
#> 456   0.78895004      0
#> 457   2.18851822      0
#> 458   0.18999810      0
#> 459   2.60070724      0
#> 460  -1.22485882      0
#> 461  -0.36899985      0
#> 462  -0.77549983      0
#> 463  -0.53388265      0
#> 464  -0.18748800      0
#> 465  -1.20573077      0
#> 466   0.71482590      0
#> 467  -0.20586969      0
#> 468   0.54510535      0
#> 469  -1.46726325      0
#> 470  -1.29927953      0
#> 471  -0.42225714      0
#> 472   0.04686012      0
#> 473  -0.78690339      0
#> 474  -0.18402353      0
#> 475   0.45068575      0
#> 476  -1.81462606      0
#> 477  -0.31708423      0
#> 478  -0.30891511      0
#> 479  -0.39637253      0
#> 480  -0.07621756      0
#> 481   0.09470262      0
#> 482   0.54598681      0
#> 483  -1.01647404      0
#> 484   0.70107294      0
#> 485  -1.33359412      0
#> 486  -0.19922699      0
#> 487   1.56740059      0
#> 488  -2.10873667      0
#> 489   1.28664439      0
#> 490   0.62000704      0
#> 491   1.21151753      0
#> 492  -0.32326213      0
#> 493  -1.79515431      0
#> 494  -0.35631134      0
#> 495  -1.33167128      0
#> 496  -0.60848128      0
#> 497   1.06874076      0
#> 498  -0.64199346      0
#> 499  -0.25501582      0
#> 500  -0.51882362      0
#> 501  -0.95645202      0
#> 502   0.78286644      0
#> 503   0.30977424      0
#> 504   1.05659310      0
#> 505  -1.33264971      0
#> 506  -0.30887019      0
#> 507  -0.22638890      0
#> 508  -0.01066035      0
#> 509   2.41052246      0
#> 510   0.32729062      0
#> 511  -0.74922555      0
#> 512  -0.31914624      0
#> 513   0.88523384      0
#> 514  -0.13943807      0
#> 515  -0.48190774      0
#> 516   1.81007310      0
#> 517  -0.16733583      0
#> 518   0.26460046      0
#> 519   0.97265865      0
#> 520   0.18262281      0
#> 521  -0.01361545      0
#> 522   1.45614573      0
#> 523   0.43118651      0
#> 524  -1.55280196      0
#> 525  -0.54960980      0
#> 526  -1.40752829      0
#> 527  -0.45745425      0
#> 528   0.60014092      0
#> 529  -0.54437247      0
#> 530  -1.27663669      0
#> 531   1.70808169      0
#> 532  -0.30221261      0
#> 533   0.80601737      0
#> 534   1.64550958      0
#> 535   0.08786308      0
#> 536  -0.06743008      0
#> 537   1.87834504      0
#> 538  -0.91841408      0
#> 539   2.05044127      0
#> 540   0.29879697      0
#> 541   2.09781766      0
#> 542  -0.13210456      0
#> 543   1.06525014      0
#> 544   0.97952612      0
#> 545   0.34138507      0
#> 546  -0.90242755      0
#> 547  -0.13231171      0
#> 548   1.06578402      0
#> 549  -0.62040394      0
#> 550  -1.75555739      0
#> 551  -1.25595097      0
#> 552  -1.15984805      0
#> 553   1.31841207      0
#> 554   2.35198099      0
#> 555  -0.66597500      0
#> 556  -0.19519052      0
#> 557  -0.11245370      0
#> 558   0.50730185      0
#> 559  -0.02304731      0
#> 560   0.79083986      0
#> 561   0.93606398      0
#> 562   0.46669586      0
#> 563   0.92041372      0
#> 564  -1.00254553      0
#> 565   0.81975983      0
#> 566   0.52234873      0
#> 567   0.73031431      0
#> 568   0.07831392      0
#> 569   1.55185209      0
#> 570  -0.50295924      0
#> 571   1.06984037      0
#> 572   0.02541805      0
#> 573  -1.22617782      0
#> 574   2.37058421      0
#> 575   0.13482858      0
#> 576  -0.16566860      0
#> 577   0.06650382      0
#> 578   1.17534113      0
#> 579   1.30914345      0
#> 580   0.67393584      0
#> 581   0.73566635      0
#> 582  -0.03617365      0
#> 583   1.06210930      0
#> 584   1.64174729      0
#> 585  -1.35449576      0
#> 586  -0.49424199      0
#> 587   1.41547686      0
#> 588   0.68966120      0
#> 589   1.06007252      0
#> 590  -1.81399891      0
#> 591  -0.10241736      0
#> 592   1.16114004      0
#> 593   1.14336608      0
#> 594  -1.94769821      0
#> 595  -0.17242749      0
#> 596  -0.31352147      0
#> 597   0.16189314      0
#> 598   0.53897209      0
#> 599  -0.65294268      0
#> 600   0.60734565      0
#> 601  -0.31875772      0
#> 602  -0.55745210      0
#> 603   0.32549353      0
#> 604   0.38321915      0
#> 605   1.18988659      0
#> 606   0.15826052      0
#> 607  -0.55979699      0
#> 608  -0.30947388      0
#> 609  -0.79629401      0
#> 610  -1.13370487      0
#> 611   1.69541254      0
#> 612   1.00596140      0
#> 613  -0.87451268      0
#> 614  -0.61937666      0
#> 615  -1.12227837      0
#> 616  -0.02622783      0
#> 617   0.96207853      0
#> 618   0.61262356      0
#> 619  -0.32799694      0
#> 620   0.89796545      0
#> 621  -0.16567880      0
#> 622   0.64726408      0
#> 623  -0.14462638      0
#> 624   2.02199170      0
#> 625  -0.75721294      0
#> 626  -0.43756427      0
#> 627  -0.84160882      0
#> 628  -1.57970588      0
#> 629   1.20277213      0
#> 630   0.81187113      0
#> 631   0.18507438      0
#> 632   1.59556822      0
#> 633   1.54235211      0
#> 634  -0.69271447      0
#> 635  -0.67743291      0
#> 636  -0.05911170      0
#> 637  -0.68185844      0
#> 638   1.35673733      0
#> 639  -0.24410931      0
#> 640  -0.88248922      0
#> 641   0.97605275      0
#> 642  -1.13526322      0
#> 643  -0.06244238      0
#> 644   1.06091405      0
#> 645  -0.65441674      0
#> 646  -1.67574441      0
#> 647   0.50320354      0
#> 648   0.51810101      0
#> 649  -0.63102376      0
#> 650  -0.72723304      0
#> 651  -0.80548159      0
#> 652  -0.61088424      0
#> 653  -0.79642009      0
#> 654   1.29836572      0
#> 655   0.06782593      0
#> 656  -0.72162906      0
#> 657   0.92514202      0
#> 658   0.68194384      0
#> 659   0.59320111      0
#> 660   0.06735240      0
#> 661  -1.99267146      0
#> 662   0.85443143      0
#> 663  -1.17266318      0
#> 664  -0.75718987      0
#> 665  -0.24278796      0
#> 666  -1.27906221      0
#> 667  -0.77402939      0
#> 668   0.19488447      0
#> 669  -0.83860248      0
#> 670  -2.78813704      0
#> 671   0.86273031      0
#> 672   1.64815256      0
#> 673  -0.87787032      0
#> 674   1.58921452      0
#> 675   0.59448810      0
#> 676  -0.82725490      0
#> 677  -1.15362881      0
#> 678   0.05546492      0
#> 679   1.18200701      0
#> 680   0.49747666      0
#> 681   0.51019892      0
#> 682   0.79307964      0
#> 683  -2.65044544      0
#> 684   3.25371892      0
#> 685  -1.08356773      0
#> 686  -0.59157140      0
#> 687  -1.50255880      0
#> 688   1.08077687      0
#> 689   1.95133635      0
#> 690   1.61592829      0
#> 691   1.98172590      0
#> 692   0.84665136      0
#> 693  -0.16012620      0
#> 694   0.85038668      0
#> 695   1.82784632      0
#> 696   0.29471188      0
#> 697  -0.05221120      0
#> 698   0.38007458      0
#> 699   0.71269305      0
#> 700   0.98899830      0
#> 701   1.21484198      0
#> 702  -1.32984318      0
#> 703  -1.44069079      0
#> 704  -0.16368086      0
#> 705  -0.10103643      0
#> 706  -2.24198080      0
#> 707   1.09819610      0
#> 708   0.02329959      0
#> 709   0.34566846      0
#> 710   1.09753456      0
#> 711   1.37835078      0
#> 712   0.95613242      0
#> 713  -0.77156857      0
#> 714  -1.47502848      0
#> 715  -0.08956860      0
#> 716   0.71896349      0
#> 717   0.52917161      0
#> 718  -1.04035492      0
#> 719   0.10935693      0
#> 720   2.08891902      0
#> 721   1.67354934      0
#> 722   0.85046945      0
#> 723   0.56177996      0
#> 724  -0.18700526      0
#> 725  -2.32501581      0
#> 726   0.41005969      0
#> 727  -0.91326076      0
#> 728   0.63426723      0
#> 729  -0.07334526      0
#> 730  -0.53540401      0
#> 731  -0.11939836      0
#> 732  -0.44891151      0
#> 733   3.10470531      0
#> 734   1.48443827      0
#> 735  -1.48993907      0
#> 736   0.14443160      0
#> 737   0.03815944      0
#> 738  -0.73557614      0
#> 739  -2.40281243      0
#> 740  -1.84510447      0
#> 741   0.23172913      0
#> 742  -0.57254845      0
#> 743  -0.65706521      0
#> 744  -1.84535705      0
#> 745  -1.37331958      0
#> 746  -0.45798589      0
#> 747  -0.76665599      0
#> 748   0.05675179      0
#> 749   0.32740026      0
#> 750  -1.51672915      0
#> 751  -0.07836550      0
#> 752   0.40124210      0
#> 753  -0.58004982      0
#> 754   2.43750834      0
#> 755   1.25793268      0
#> 756   0.93970010      0
#> 757  -0.99994859      0
#> 758  -1.74834131      0
#> 759  -1.86610791      0
#> 760   0.18335954      0
#> 761  -0.94811793      0
#> 762  -0.13173316      0
#> 763  -0.48134227      0
#> 764  -1.38708710      0
#> 765  -1.19945768      0
#> 766  -0.80927986      0
#> 767  -1.27241377      0
#> 768  -0.19132204      0
#> 769  -1.10404908      0
#> 770  -0.68178540      0
#> 771   0.53827131      0
#> 772  -2.81226589      0
#> 773  -0.69342820      0
#> 774   0.75351393      0
#> 775   1.03544090      0
#> 776   0.97937632      0
#> 777   1.07395070      0
#> 778   1.45744763      0
#> 779   0.56285572      0
#> 780   0.28986234      0
#> 781   1.71516696      0
#> 782  -0.49236712      0
#> 783  -0.30347905      0
#> 784  -1.24441174      0
#> 785  -0.48131461      0
#> 786   1.06529618      0
#> 787  -2.53543624      0
#> 788  -0.35082816      0
#> 789   0.29466782      0
#> 790   0.06720236      0
#> 791   0.41242818      0
#> 792  -0.57619535      0
#> 793  -0.02213514      0
#> 794   1.10894789      0
#> 795  -0.14167165      0
#> 796   1.86164683      0
#> 797   0.68665941      0
#> 798   1.49466226      0
#> 799  -1.35784699      0
#> 800  -0.26830090      0
#> 801   1.10911771      0
#> 802  -0.47663152      0
#> 803   0.81504522      0
#> 804   0.50798094      0
#> 805   0.52458022      0
#> 806   0.15876262      0
#> 807   0.87715959      0
#> 808   0.20628426      0
#> 809  -0.51655315      0
#> 810  -1.10982909      0
#> 811   2.33409574      0
#> 812  -1.30710429      0
#> 813   0.68213768      0
#> 814  -0.34130816      0
#> 815  -0.64354764      0
#> 816  -0.69453459      0
#> 817   2.25087373      0
#> 818   0.25385515      0
#> 819   1.16668358      0
#> 820   0.36827114      0
#> 821   1.19493756      0
#> 822  -0.15115692      0
#> 823   2.24693278      0
#> 824  -0.10924696      0
#> 825  -0.36173113      0
#> 826  -0.74782308      0
#> 827  -0.69464081      0
#> 828  -0.32023608      0
#> 829   0.04703121      0
#> 830  -0.46717942      0
#> 831   0.61127942      0
#> 832  -0.16831479      0
#> 833   0.78406674      0
#> 834  -0.07164462      0
#> 835   1.22098763      0
#> 836   0.69740067      0
#> 837   0.31648561      0
#> 838  -0.22920514      0
#> 839  -1.04036814      0
#> 840   1.18692114      0
#> 841   1.74165803      0
#> 842   1.66994053      0
#> 843  -0.55741266      0
#> 844   1.61086265      0
#> 845  -0.56475256      0
#> 846  -1.74018321      0
#> 847   0.54197306      0
#> 848   0.18456906      0
#> 849  -0.74950094      0
#> 850   0.74274963      0
#> 851  -1.27903946      0
#> 852   1.53814485      0
#> 853   1.72639809      0
#> 854  -0.75684887      0
#> 855  -1.38572617      0
#> 856  -0.51533122      0
#> 857   2.34691884      0
#> 858  -0.72823835      0
#> 859   1.55885581      0
#> 860  -1.35315572      0
#> 861  -1.84455153      0
#> 862   0.88684037      0
#> 863   1.42183275      0
#> 864   0.85105259      0
#> 865  -2.13219705      0
#> 866  -1.49514558      0
#> 867  -0.73662844      0
#> 868  -1.72131588      0
#> 869   0.42017998      0
#> 870  -1.33392374      0
#> 871  -1.57255184      0
#> 872   0.20881205      0
#> 873  -0.09490580      0
#> 874  -0.20167671      0
#> 875   0.63702480      0
#> 876  -0.18359838      0
#> 877   1.30502336      0
#> 878   2.50615904      0
#> 879  -1.00751620      0
#> 880  -0.75622841      0
#> 881  -1.72515816      0
#> 882  -0.12381420      0
#> 883  -1.60560412      0
#> 884   1.58512808      0
#> 885   0.11632395      0
#> 886   0.96482001      0
#> 887  -0.75312160      0
#> 888   0.43201438      0
#> 889  -0.10733337      0
#> 890   1.27507247      0
#> 891   1.41218743      0
#> 892  -1.65099164      0
#> 893  -0.59737664      0
#> 894   0.69939440      0
#> 895   1.75044319      0
#> 896  -0.04707582      0
#> 897   0.89948652      0
#> 898   0.23127356      0
#> 899   0.47470991      1
#> 900   1.38415182      1
#> 901   0.13358003      1
#> 902   1.46812342      1
#> 903   1.32894554      1
#> 904  -0.93554731      1
#> 905   1.49534167      1
#> 906   1.44183444      1
#> 907   1.02414655      1
#> 908   0.28572963      1
#> 909   1.53515670      1
#> 910   0.12707390      1
#> 911   0.76941032      1
#> 912  -0.79100797      1
#> 913   0.69624530      1
#> 914   2.28836264      1
#> 915   0.13428474      1
#> 916   0.53023505      1
#> 917  -0.57720340      1
#> 918   0.74857796      1
#> 919   1.69011603      1
#> 920   1.62846112      1
#> 921   1.06218769      1
#> 922   0.17900454      1
#> 923   0.11057014      1
#> 924  -0.22984449      1
#> 925  -0.78317328      1
#> 926   1.38546462      1
#> 927  -0.43379513      1
#> 928   2.13517968      1
#> 929  -1.32643199      1
#> 930   0.82509097      1
#> 931   1.09064501      1
#> 932   0.49775970      1
#> 933   0.47615350      1
#> 934   0.63034227      1
#> 935   0.40754151      1
#> 936  -0.56879357      1
#> 937   0.56866970      1
#> 938   2.66350278      1
#> 939  -0.28884457      1
#> 940  -0.23141808      1
#> 941   1.32767054      1
#> 942   1.18776278      1
#> 943  -0.33350273      1
#> 944   1.34505266      1
#> 945   1.67244255      1
#> 946   1.71952041      1
#> 947   0.59042907      1
#> 948   0.95506268      1
#> 949  -0.28981144      1
#> 950   0.87953926      1
#> 951   0.14943831      1
#> 952   0.66579617      1
#> 953   0.71107567      1
#> 954   1.20153084      1
#> 955  -0.23490917      1
#> 956   1.88896792      1
#> 957   1.64551670      1
#> 958   0.47032383      1
#> 959  -0.96296391      1
#> 960   1.40485893      1
#> 961   1.65350242      1
#> 962   1.02541764      1
#> 963  -0.10671461      1
#> 964   1.65483298      1
#> 965  -0.96452874      1
#> 966   2.72219585      1
#> 967   1.94655693      1
#> 968   0.11144333      1
#> 969   0.67697640      1
#> 970   0.48734556      1
#> 971   0.45049608      1
#> 972  -0.29260669      1
#> 973   0.13182630      1
#> 974   0.85392519      1
#> 975   0.11869020      1
#> 976   1.02070344      1
#> 977   0.42140470      1
#> 978   1.09332685      1
#> 979   0.37501508      1
#> 980   2.00295681      1
#> 981   0.32198260      1
#> 982   2.22251657      1
#> 983   2.17753984      1
#> 984   3.05474385      1
#> 985   3.09476265      1
#> 986  -0.19014368      1
#> 987   1.38499441      1
#> 988   2.00399466      1
#> 989   1.45291763      1
#> 990   0.56571568      1
#> 991   1.03313911      1
#> 992   0.49243272      1
#> 993   1.98301544      1
#> 994   1.79891376      1
#> 995  -0.09648324      1
#> 996   0.72689318      1
#> 997   0.14229431      1
#> 998   2.08040386      1
#> 999   1.76233918      1
#> 1000  1.38105956      1