Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.825582173      0
#> 2     1.633276875      0
#> 3    -0.358352772      0
#> 4    -0.273226412      0
#> 5     1.249452255      0
#> 6     0.235259202      0
#> 7     0.106668151      0
#> 8    -1.061418309      0
#> 9    -0.732428003      0
#> 10   -1.272155634      0
#> 11   -0.548640325      0
#> 12    0.556935131      0
#> 13    1.067862020      0
#> 14    0.070652389      0
#> 15    0.934143158      0
#> 16    0.703183845      0
#> 17   -0.964542112      0
#> 18   -0.459482706      0
#> 19   -0.932286439      0
#> 20   -0.763362915      0
#> 21   -0.314381504      0
#> 22    0.244505797      0
#> 23   -1.361028953      0
#> 24    0.885513582      0
#> 25    0.541429971      0
#> 26   -0.998790840      0
#> 27   -0.205524702      0
#> 28    0.441881376      0
#> 29    0.996278034      0
#> 30    0.107808169      0
#> 31    0.813453404      0
#> 32   -0.561872112      0
#> 33    0.200056307      0
#> 34    0.173717558      0
#> 35   -3.050313125      0
#> 36    1.085484782      0
#> 37   -0.949154756      0
#> 38   -0.286433342      0
#> 39   -0.470059838      0
#> 40    0.878115017      0
#> 41   -0.870290457      0
#> 42    1.309921603      0
#> 43    2.107776185      0
#> 44   -0.598364384      0
#> 45    1.892988032      0
#> 46   -0.285953837      0
#> 47    0.890266165      0
#> 48   -0.812769387      0
#> 49    1.295368511      0
#> 50   -0.929279643      0
#> 51   -1.594162968      0
#> 52    0.553150291      0
#> 53   -0.596609610      0
#> 54   -0.001534926      0
#> 55   -0.445018559      0
#> 56   -0.508615420      0
#> 57   -0.747334632      0
#> 58   -0.067670324      0
#> 59   -1.567892953      0
#> 60   -1.580680839      0
#> 61   -0.753821343      0
#> 62   -2.354766791      0
#> 63   -0.167908568      0
#> 64   -0.793351513      0
#> 65   -0.132791875      0
#> 66    0.489819791      0
#> 67   -0.579116309      0
#> 68   -0.126943721      0
#> 69    1.371322028      0
#> 70    0.313901396      0
#> 71    0.520049030      0
#> 72   -0.838016651      0
#> 73   -0.184574154      0
#> 74    1.477996030      0
#> 75   -0.048107170      0
#> 76    0.828425522      0
#> 77   -0.353484798      0
#> 78   -0.864976110      0
#> 79    0.481081721      0
#> 80   -0.230358533      0
#> 81    0.533584661      0
#> 82    0.479975594      0
#> 83   -1.713403058      0
#> 84    0.821584433      0
#> 85    1.286728084      0
#> 86    1.357396829      0
#> 87   -0.451035779      0
#> 88   -2.198890095      0
#> 89   -0.891785277      0
#> 90   -1.705318865      0
#> 91   -0.264537691      0
#> 92   -0.725458711      0
#> 93   -0.007485846      0
#> 94    0.970454827      0
#> 95    0.680704374      0
#> 96    1.022012804      0
#> 97   -0.125364486      0
#> 98    0.578401813      0
#> 99   -0.907984539      0
#> 100  -0.454211635      0
#> 101  -1.005669292      0
#> 102  -0.695319817      0
#> 103  -2.283396041      0
#> 104  -1.561977517      0
#> 105   1.339420796      0
#> 106   1.198585940      0
#> 107   0.159790199      0
#> 108  -1.302349411      0
#> 109  -0.245547022      0
#> 110  -0.875107298      0
#> 111  -0.801841705      0
#> 112   0.295764382      0
#> 113   0.083959730      0
#> 114  -0.845307738      0
#> 115  -0.141754541      0
#> 116   0.439582655      0
#> 117   0.428969784      0
#> 118   0.455559377      0
#> 119   2.221232389      0
#> 120   0.727214574      0
#> 121  -0.745025531      0
#> 122  -1.242647635      0
#> 123  -1.624989989      0
#> 124   0.027698360      0
#> 125   0.475885302      0
#> 126  -1.106370802      0
#> 127   0.742626676      0
#> 128  -0.252507673      0
#> 129   1.134723510      0
#> 130   0.311774635      0
#> 131   0.903285794      0
#> 132  -2.304225571      0
#> 133   0.294127195      0
#> 134   1.312202419      0
#> 135  -0.035663414      0
#> 136  -0.362348712      0
#> 137   0.955781269      0
#> 138   1.273682054      0
#> 139   1.072627589      0
#> 140  -1.305028778      0
#> 141  -0.724651736      0
#> 142   0.551435984      0
#> 143  -0.417448130      0
#> 144  -2.618242013      0
#> 145  -0.811346891      0
#> 146   0.005952290      0
#> 147  -2.080367041      0
#> 148   0.367956563      0
#> 149  -0.738629101      0
#> 150   0.079728335      0
#> 151  -0.790098235      0
#> 152  -3.437408267      0
#> 153  -0.841265076      0
#> 154   1.318682377      0
#> 155  -0.409158987      0
#> 156   0.481216378      0
#> 157  -0.589854266      0
#> 158   0.314173676      0
#> 159  -0.033930966      0
#> 160  -0.047926333      0
#> 161   0.111467385      0
#> 162   1.893420627      0
#> 163   0.280870213      0
#> 164   0.760878225      0
#> 165  -2.169430407      0
#> 166   1.122234550      0
#> 167   0.073829333      0
#> 168  -0.110465591      0
#> 169   1.824296572      0
#> 170  -1.246411618      0
#> 171   0.958761954      0
#> 172  -0.942291369      0
#> 173  -1.787079556      0
#> 174  -0.432860241      0
#> 175  -0.660612383      0
#> 176   0.689025240      0
#> 177   0.410664090      0
#> 178  -1.152213978      0
#> 179   1.702604873      0
#> 180   1.375129022      0
#> 181   1.713491046      0
#> 182  -1.651220063      0
#> 183   0.913194536      0
#> 184   2.308098056      0
#> 185   2.132036844      0
#> 186  -0.273010563      0
#> 187  -0.411356112      0
#> 188  -1.175890798      0
#> 189  -0.625506088      0
#> 190   0.958279835      0
#> 191  -2.316171322      0
#> 192  -0.346419097      0
#> 193   1.548885108      0
#> 194  -0.315376266      0
#> 195  -0.192354010      0
#> 196  -2.307138602      0
#> 197  -0.076307315      0
#> 198   0.072853390      0
#> 199   1.147454449      0
#> 200   1.182508924      0
#> 201   0.493726605      0
#> 202  -0.090196808      0
#> 203  -1.059843365      0
#> 204  -0.398338181      0
#> 205  -0.313079396      0
#> 206   1.027781150      0
#> 207  -1.053791756      0
#> 208   0.700536197      0
#> 209  -1.426418903      0
#> 210  -0.356611302      0
#> 211   1.103361976      0
#> 212   1.380749796      0
#> 213   1.690613128      0
#> 214  -0.422666402      0
#> 215  -0.024131885      0
#> 216   0.150876660      0
#> 217  -1.691092045      0
#> 218   0.982067785      0
#> 219   0.832054651      0
#> 220   0.967836346      0
#> 221  -0.023147464      0
#> 222  -0.175424480      0
#> 223   0.240567235      0
#> 224  -0.890328773      0
#> 225   1.017127271      0
#> 226  -0.874702544      0
#> 227  -0.090139156      0
#> 228   0.207746125      0
#> 229  -0.856619770      0
#> 230  -1.674115902      0
#> 231   1.532102304      0
#> 232   1.347608382      0
#> 233  -2.041509831      0
#> 234   2.153073800      0
#> 235  -0.409283632      0
#> 236   1.458551378      0
#> 237   0.878572990      0
#> 238   0.701600807      0
#> 239   0.319130922      0
#> 240   0.709437694      0
#> 241  -0.472004375      0
#> 242  -1.552042295      0
#> 243  -0.037666618      0
#> 244   0.951624628      0
#> 245   0.728186076      0
#> 246  -0.629880561      0
#> 247   0.923467379      0
#> 248   1.092318443      0
#> 249   1.223024777      0
#> 250  -1.315974109      0
#> 251   0.224310424      0
#> 252   2.019291990      0
#> 253  -1.863006484      0
#> 254   0.589974312      0
#> 255   0.709292943      0
#> 256  -0.677373449      0
#> 257  -1.145019596      0
#> 258  -0.561354116      0
#> 259  -1.078405756      0
#> 260   0.384709103      0
#> 261  -1.623156989      0
#> 262   0.194258323      0
#> 263   1.882108110      0
#> 264  -0.601046868      0
#> 265  -1.108220112      0
#> 266   0.013230185      0
#> 267  -0.194812473      0
#> 268   2.253232291      0
#> 269   1.420897787      0
#> 270   0.514723835      0
#> 271  -0.451558437      0
#> 272  -3.502263823      0
#> 273  -0.171881602      0
#> 274  -0.570796534      0
#> 275  -0.300340981      0
#> 276  -0.667792406      0
#> 277  -1.031492148      0
#> 278   1.063031502      0
#> 279   0.680447124      0
#> 280   0.183810013      0
#> 281   0.578672294      0
#> 282  -0.930752000      0
#> 283  -0.333480121      0
#> 284   0.760246782      0
#> 285  -2.040840255      0
#> 286  -0.036427977      0
#> 287   1.302760854      0
#> 288   1.272159057      0
#> 289  -0.606389901      0
#> 290  -0.313965853      0
#> 291   1.017205792      0
#> 292  -1.255687686      0
#> 293  -0.312224542      0
#> 294  -2.120208879      0
#> 295   0.285327183      0
#> 296   1.829558959      0
#> 297  -0.843036754      0
#> 298   2.589496119      0
#> 299  -1.566311268      0
#> 300  -0.255200499      0
#> 301  -0.350169933      0
#> 302   0.628700936      0
#> 303   0.309967135      0
#> 304   0.222044360      0
#> 305   1.623914198      0
#> 306  -0.317037304      0
#> 307  -1.934477839      0
#> 308   0.481531012      0
#> 309   0.959597375      0
#> 310   0.807225246      0
#> 311   0.271839858      0
#> 312   1.362979602      0
#> 313  -2.297155955      0
#> 314   2.026917091      0
#> 315   1.203108418      0
#> 316   3.430498218      0
#> 317  -0.685971632      0
#> 318  -0.911197278      0
#> 319   1.579554308      0
#> 320  -0.882758887      0
#> 321   0.750987837      0
#> 322  -0.588665026      0
#> 323  -0.596147129      0
#> 324  -2.092656716      0
#> 325  -1.503571877      0
#> 326  -1.156086882      0
#> 327   0.718510112      0
#> 328  -0.781103733      0
#> 329  -1.999494379      0
#> 330  -2.145606248      0
#> 331   2.837470077      0
#> 332   2.064372646      0
#> 333  -1.475980479      0
#> 334   0.201281497      0
#> 335   0.023459381      0
#> 336   1.236793885      0
#> 337  -1.213356251      0
#> 338  -2.833490641      0
#> 339  -0.920480914      0
#> 340  -0.427936201      0
#> 341   0.040753640      0
#> 342   1.253112778      0
#> 343  -1.118671669      0
#> 344   0.231733657      0
#> 345  -0.423895925      0
#> 346   0.589575677      0
#> 347  -0.048348135      0
#> 348  -0.226768945      0
#> 349   0.063587203      0
#> 350   0.218694209      0
#> 351  -2.149580052      0
#> 352  -0.759331130      0
#> 353  -0.542148851      0
#> 354   0.848863807      0
#> 355   0.776485190      0
#> 356   0.321693200      0
#> 357   0.421227577      0
#> 358  -0.529109645      0
#> 359  -1.729762254      0
#> 360  -1.031876703      0
#> 361  -0.406745956      0
#> 362   0.533914943      0
#> 363   0.027507431      0
#> 364  -0.997041895      0
#> 365  -1.002728953      0
#> 366  -1.122876365      0
#> 367   0.888813740      0
#> 368   0.048035009      0
#> 369  -0.230831761      0
#> 370  -0.555768192      0
#> 371   0.594006284      0
#> 372   1.158701458      0
#> 373  -0.203687983      0
#> 374   0.003746125      0
#> 375   0.488996461      0
#> 376  -1.415006930      0
#> 377   1.593398053      0
#> 378   0.266980819      0
#> 379   0.223197151      0
#> 380  -0.492708057      0
#> 381   1.020276759      0
#> 382  -1.106589887      0
#> 383  -0.428396271      0
#> 384   1.665396807      0
#> 385   0.403584052      0
#> 386  -1.071558162      0
#> 387   1.391481951      0
#> 388   0.250804150      0
#> 389   0.723507581      0
#> 390   2.446306654      0
#> 391   1.242443044      0
#> 392   0.592619621      0
#> 393   0.536353077      0
#> 394   0.562495538      0
#> 395   0.558838607      0
#> 396  -0.237835344      0
#> 397  -0.723072264      0
#> 398  -0.667692208      0
#> 399   0.550689602      0
#> 400  -0.742130856      0
#> 401  -0.739738527      0
#> 402   0.528318540      0
#> 403  -0.731109023      0
#> 404  -0.096060608      0
#> 405  -1.180008582      0
#> 406   0.397055449      0
#> 407   0.297456710      0
#> 408  -2.016757893      0
#> 409   0.014661385      0
#> 410   0.810506479      0
#> 411  -0.572216484      0
#> 412  -1.158504958      0
#> 413   0.581289330      0
#> 414  -0.270649649      0
#> 415   1.818775742      0
#> 416   1.018135059      0
#> 417   1.633572407      0
#> 418  -0.839450312      0
#> 419   0.606838841      0
#> 420  -0.859158035      0
#> 421  -0.650824303      0
#> 422   0.393034566      0
#> 423   0.985880727      0
#> 424   0.173245069      0
#> 425  -0.861283021      0
#> 426  -1.170350285      0
#> 427   0.761998810      0
#> 428  -1.052134945      0
#> 429   0.958623092      0
#> 430  -0.036735755      0
#> 431  -0.521327923      0
#> 432  -0.523158444      0
#> 433   1.059998258      0
#> 434  -0.127217681      0
#> 435   0.679255631      0
#> 436   0.518086554      0
#> 437  -0.852077779      0
#> 438   0.992715321      0
#> 439  -0.822254517      0
#> 440  -1.109723775      0
#> 441   0.141103330      0
#> 442   0.381551148      0
#> 443  -0.722965356      0
#> 444  -0.877854448      0
#> 445   0.665514733      0
#> 446   1.448521557      0
#> 447  -1.829473235      0
#> 448   0.555035407      0
#> 449  -0.592331157      0
#> 450  -1.161490941      0
#> 451   0.181170433      0
#> 452   0.017036701      0
#> 453   0.995019363      0
#> 454  -2.451137561      0
#> 455  -0.068789043      0
#> 456  -0.979627173      0
#> 457   0.858805643      0
#> 458   0.358071425      0
#> 459  -0.763962694      0
#> 460  -0.264322190      0
#> 461   1.214746617      0
#> 462   0.380932218      0
#> 463  -0.092617157      0
#> 464   0.412742235      0
#> 465  -0.736575752      0
#> 466  -1.170615964      0
#> 467   0.587441318      0
#> 468  -0.808131912      0
#> 469  -0.317598932      0
#> 470   0.421293029      0
#> 471  -0.964952124      0
#> 472   0.782190738      0
#> 473   0.127423673      0
#> 474  -0.615673640      0
#> 475  -1.523798625      0
#> 476  -0.235017109      0
#> 477  -0.436306353      0
#> 478  -0.806054741      0
#> 479   0.690007262      0
#> 480   0.836270796      0
#> 481  -0.845520337      0
#> 482  -1.117459124      0
#> 483  -0.303380201      0
#> 484  -1.050439816      0
#> 485   1.613826521      0
#> 486  -1.339593769      0
#> 487  -1.193385161      0
#> 488  -1.808901688      0
#> 489  -0.310530627      0
#> 490   0.147861496      0
#> 491   0.805352171      0
#> 492   0.239596662      0
#> 493   0.256481044      0
#> 494   0.327352370      0
#> 495   0.643624711      0
#> 496   0.905140668      0
#> 497   1.385767288      0
#> 498   0.893555445      0
#> 499   1.075678192      0
#> 500  -0.330446258      0
#> 501   1.381769753      0
#> 502  -0.443763343      0
#> 503  -0.279718980      0
#> 504  -1.154520912      0
#> 505  -2.507873842      0
#> 506  -0.735529477      0
#> 507   0.020867813      0
#> 508  -1.756214118      0
#> 509  -0.793551689      0
#> 510   0.649023074      0
#> 511   0.614362167      0
#> 512  -0.684536083      0
#> 513  -0.169136875      0
#> 514   1.482700516      0
#> 515  -0.498983483      0
#> 516  -0.012834779      0
#> 517   0.215285127      0
#> 518  -1.228208649      0
#> 519   0.069062235      0
#> 520   0.630133242      0
#> 521  -0.788040841      0
#> 522  -1.175913325      0
#> 523  -1.807081020      0
#> 524   1.118874022      0
#> 525  -0.875620022      0
#> 526   0.315256103      0
#> 527   0.467203017      0
#> 528   0.990071708      0
#> 529   0.304146933      0
#> 530   1.155469442      0
#> 531  -0.904381255      0
#> 532   0.079352835      0
#> 533  -0.485194666      0
#> 534  -0.332574615      0
#> 535  -1.186576012      0
#> 536   0.555654107      0
#> 537  -0.935891063      0
#> 538   1.159516733      0
#> 539  -0.208718308      0
#> 540   0.614101038      0
#> 541   1.242359115      0
#> 542  -1.140908453      0
#> 543   0.248807388      0
#> 544  -0.399575639      0
#> 545  -2.125618555      0
#> 546   0.859207359      0
#> 547  -0.027104713      0
#> 548   0.864173474      0
#> 549   0.417185548      0
#> 550  -0.460786496      0
#> 551   0.070499289      0
#> 552  -0.824708739      0
#> 553   0.669997157      0
#> 554  -0.203758468      0
#> 555  -0.459922899      0
#> 556  -0.672308634      0
#> 557  -0.104316951      0
#> 558   0.898959647      0
#> 559  -1.092306903      0
#> 560  -1.240407286      0
#> 561  -0.715494830      0
#> 562   0.380435505      0
#> 563  -1.359796803      0
#> 564  -2.675038487      0
#> 565   0.286859697      0
#> 566  -0.492078029      0
#> 567  -1.953882033      0
#> 568   0.564743958      0
#> 569   1.227525082      0
#> 570  -1.286929197      0
#> 571  -1.095382539      0
#> 572  -1.145851858      0
#> 573   0.364001799      0
#> 574   0.175256649      0
#> 575  -1.523043104      0
#> 576   1.683792776      0
#> 577  -0.082447708      0
#> 578  -0.467876498      0
#> 579  -1.043842256      0
#> 580   0.657893289      0
#> 581  -0.824526765      0
#> 582  -0.595240961      0
#> 583  -0.409530979      0
#> 584   0.059091101      0
#> 585  -0.260387996      0
#> 586   0.421300319      0
#> 587  -0.256878605      0
#> 588  -0.138195534      0
#> 589   0.571162536      0
#> 590  -0.834816825      0
#> 591   0.085057021      0
#> 592   0.842947037      0
#> 593   1.108400500      0
#> 594  -0.392053947      0
#> 595   0.698733908      0
#> 596  -0.442735259      0
#> 597  -1.505696613      0
#> 598  -0.528532656      0
#> 599   0.267272343      0
#> 600  -0.684447702      0
#> 601   0.037143793      0
#> 602   0.381517114      0
#> 603   0.907648869      0
#> 604  -0.194272487      0
#> 605   0.595814678      0
#> 606   0.210001196      0
#> 607   1.047721705      0
#> 608   1.607420477      0
#> 609   0.393850060      0
#> 610  -0.231069083      0
#> 611  -0.091497361      0
#> 612   1.317535847      0
#> 613  -1.157533405      0
#> 614  -1.119258835      0
#> 615   1.619567645      0
#> 616  -0.668106618      0
#> 617  -0.882276421      0
#> 618   1.140070228      0
#> 619  -0.185643831      0
#> 620  -0.191126368      0
#> 621  -0.634500428      0
#> 622  -0.425369063      0
#> 623  -0.347147152      0
#> 624  -0.561153074      0
#> 625  -1.001072067      0
#> 626   0.093832344      0
#> 627  -0.600535535      0
#> 628  -1.101245434      0
#> 629   0.271111946      0
#> 630   0.915416691      0
#> 631   0.340621160      0
#> 632  -0.327323642      0
#> 633   1.401255462      0
#> 634   0.665481731      0
#> 635  -0.659025706      0
#> 636  -1.553743242      0
#> 637   0.964079181      0
#> 638   1.371322044      0
#> 639   0.301862279      0
#> 640   0.195171334      0
#> 641   0.168093467      0
#> 642  -0.104653148      0
#> 643  -0.347086838      0
#> 644  -0.088334700      0
#> 645   1.743442015      0
#> 646   0.246684495      0
#> 647  -0.170807067      0
#> 648  -0.167483470      0
#> 649  -1.245438158      0
#> 650   1.016744677      0
#> 651  -0.111448914      0
#> 652  -0.108878514      0
#> 653   0.177359587      0
#> 654  -1.703733955      0
#> 655   0.189414742      0
#> 656  -1.575407066      0
#> 657  -0.518483062      0
#> 658  -2.175116108      0
#> 659  -1.770902264      0
#> 660   2.675183600      0
#> 661  -0.484853611      0
#> 662  -0.840435333      0
#> 663   1.205108578      0
#> 664  -0.823322331      0
#> 665  -1.089507998      0
#> 666  -0.178608271      0
#> 667   1.236490833      0
#> 668   3.158430514      0
#> 669  -1.991042971      0
#> 670  -1.189861500      0
#> 671  -2.097757648      0
#> 672  -0.047987254      0
#> 673   1.679189967      0
#> 674   0.088228035      0
#> 675   0.569020138      0
#> 676   0.004756665      0
#> 677   0.264751771      0
#> 678  -0.863436427      0
#> 679  -1.548458632      0
#> 680   0.678011578      0
#> 681   0.342052168      0
#> 682   1.505910977      0
#> 683  -0.719408564      0
#> 684  -1.667770726      0
#> 685   0.474936647      0
#> 686  -0.602516086      0
#> 687  -1.144586168      0
#> 688  -1.679468622      0
#> 689  -0.094311194      0
#> 690   2.218431910      0
#> 691  -1.515653474      0
#> 692   0.655936178      0
#> 693  -0.276070469      0
#> 694  -2.272738204      0
#> 695  -1.214134575      0
#> 696   0.142138622      0
#> 697  -1.036705805      0
#> 698   1.504078541      0
#> 699   0.237412892      0
#> 700  -1.079230800      0
#> 701   0.673458038      0
#> 702  -0.728320224      0
#> 703   0.399085902      0
#> 704  -0.695360033      0
#> 705  -1.382534041      0
#> 706  -1.119456544      0
#> 707   1.558189531      0
#> 708   1.101984235      0
#> 709  -0.265157615      0
#> 710  -1.211057833      0
#> 711   0.591020688      0
#> 712   1.444910442      0
#> 713   0.782118345      0
#> 714   0.238402254      0
#> 715   0.089168929      0
#> 716  -1.646592295      0
#> 717   0.778167254      0
#> 718   0.004414079      0
#> 719  -0.777828691      0
#> 720   0.404783440      0
#> 721   0.335682178      0
#> 722  -0.118257276      0
#> 723   1.043482131      0
#> 724   2.611477682      0
#> 725  -0.316868239      0
#> 726  -0.429556978      0
#> 727  -0.364258750      0
#> 728  -0.028067017      0
#> 729   0.558215407      0
#> 730   0.495793832      0
#> 731   0.720778254      0
#> 732  -1.176590282      0
#> 733  -0.292724275      0
#> 734  -0.360988280      0
#> 735  -1.542344401      0
#> 736  -0.303361427      0
#> 737  -1.300917259      0
#> 738  -0.710406556      0
#> 739   0.562287295      0
#> 740  -0.074744818      0
#> 741   1.512385100      0
#> 742   0.893923239      0
#> 743  -1.799092272      0
#> 744   1.182297037      0
#> 745  -0.572390815      0
#> 746  -1.430581490      0
#> 747  -0.679228504      0
#> 748  -2.044020749      0
#> 749  -1.287940078      0
#> 750  -0.372986875      0
#> 751  -0.602188894      0
#> 752   0.654641415      0
#> 753   0.199359329      0
#> 754   0.991387333      0
#> 755  -0.988605775      0
#> 756   0.356815004      0
#> 757  -0.176726834      0
#> 758   0.668906677      0
#> 759   0.485595683      0
#> 760  -0.274043600      0
#> 761   0.511327551      0
#> 762   1.376121471      0
#> 763  -1.726749043      0
#> 764   0.031823533      0
#> 765  -0.401524201      0
#> 766   0.845939224      0
#> 767   0.226221222      0
#> 768  -1.123668529      0
#> 769  -0.518204811      0
#> 770  -0.124213846      0
#> 771  -0.292769669      0
#> 772  -1.125051816      0
#> 773  -0.093752682      0
#> 774   0.495194369      0
#> 775  -1.171636686      0
#> 776  -0.267190204      0
#> 777   2.441992466      0
#> 778  -0.397828854      0
#> 779   0.561748522      0
#> 780   0.277044219      0
#> 781   0.808179291      0
#> 782  -0.361186648      0
#> 783   0.007300547      0
#> 784  -1.014246468      0
#> 785   1.419303747      0
#> 786   1.139891900      0
#> 787   0.970811158      0
#> 788   0.762123645      0
#> 789  -0.049428033      0
#> 790  -1.444708452      0
#> 791   0.044674193      0
#> 792  -1.950000323      0
#> 793  -0.168617964      0
#> 794   0.411599037      0
#> 795   0.110767405      0
#> 796  -1.310733238      0
#> 797  -0.471745692      0
#> 798  -0.361320552      0
#> 799  -0.364795308      0
#> 800  -0.529256711      0
#> 801  -0.408633568      0
#> 802   0.259298004      0
#> 803  -0.856679857      0
#> 804   0.554988530      0
#> 805  -1.207425739      0
#> 806  -1.862121962      0
#> 807  -0.693811104      0
#> 808   0.934535970      0
#> 809   1.433490664      0
#> 810   0.472629369      0
#> 811  -0.873805024      0
#> 812   1.953328578      0
#> 813   0.194581134      0
#> 814   1.816858494      0
#> 815   1.791614527      0
#> 816   0.037855750      0
#> 817  -0.005193412      0
#> 818  -0.949229764      0
#> 819  -0.266337555      0
#> 820   1.826306152      0
#> 821  -0.606485977      0
#> 822   0.873516728      0
#> 823  -1.497767018      0
#> 824   1.032859247      0
#> 825   1.896611566      0
#> 826   0.145204640      0
#> 827   0.541008458      0
#> 828  -0.642828715      0
#> 829   0.339341617      0
#> 830  -0.705409523      0
#> 831   0.343927500      0
#> 832   0.604593442      0
#> 833   1.153614887      0
#> 834   1.973668708      0
#> 835  -0.651700480      0
#> 836  -0.758517074      0
#> 837   0.645554340      0
#> 838  -0.107985873      0
#> 839  -1.106513690      0
#> 840   0.624491315      0
#> 841   0.321310887      0
#> 842   1.097141065      0
#> 843   0.120874281      0
#> 844  -0.800930851      0
#> 845   0.806792823      0
#> 846   1.340106746      0
#> 847  -0.558961788      0
#> 848  -1.184480003      0
#> 849  -0.150846797      0
#> 850   1.208934089      0
#> 851  -1.206113981      0
#> 852   0.160386309      0
#> 853   1.797874195      0
#> 854   0.251411912      0
#> 855   0.220370884      0
#> 856  -1.020045003      0
#> 857   0.048653680      0
#> 858  -1.366192047      0
#> 859   0.182404955      0
#> 860   3.219114635      0
#> 861   0.782137956      0
#> 862   2.996859707      0
#> 863  -0.081224753      0
#> 864  -0.135059116      0
#> 865  -0.304436469      0
#> 866  -0.119336769      0
#> 867   0.216282893      0
#> 868  -2.632867524      0
#> 869   0.579350346      0
#> 870  -1.005122341      0
#> 871   0.798809577      0
#> 872   0.677858414      0
#> 873  -0.567726185      0
#> 874  -0.087569203      0
#> 875  -1.198230383      0
#> 876   0.175546703      0
#> 877  -0.523726453      0
#> 878  -0.951239533      0
#> 879   0.778504616      0
#> 880   0.542865003      0
#> 881   0.187602936      0
#> 882  -1.008381232      0
#> 883   0.567319918      0
#> 884   1.793739503      0
#> 885  -0.905528395      0
#> 886   0.924092663      0
#> 887   0.471029088      0
#> 888   0.700022517      0
#> 889  -3.190555977      0
#> 890   0.714759608      0
#> 891  -0.088459438      0
#> 892   0.873643916      0
#> 893   0.614881900      0
#> 894   1.208631451      0
#> 895   0.768024337      0
#> 896   0.252738953      1
#> 897   1.036143189      1
#> 898   0.960917424      1
#> 899   1.335468788      1
#> 900   1.994971487      1
#> 901   0.268895360      1
#> 902  -0.885685660      1
#> 903   1.282672149      1
#> 904   2.074166466      1
#> 905  -0.576365970      1
#> 906   0.854379730      1
#> 907   1.207200167      1
#> 908   1.977166089      1
#> 909   1.516930938      1
#> 910   0.115145072      1
#> 911   0.713082742      1
#> 912   1.221264198      1
#> 913   0.214043013      1
#> 914   0.869754299      1
#> 915   1.808351411      1
#> 916   0.203432174      1
#> 917   1.214350920      1
#> 918   1.834794850      1
#> 919   1.096616307      1
#> 920   0.779788636      1
#> 921   0.781854942      1
#> 922  -0.186355606      1
#> 923   2.136198448      1
#> 924   2.070494477      1
#> 925   0.794046391      1
#> 926   0.324151988      1
#> 927   2.100288723      1
#> 928   0.562029741      1
#> 929   1.345450435      1
#> 930  -0.148867317      1
#> 931  -0.076515464      1
#> 932   2.671125642      1
#> 933   1.209167369      1
#> 934   1.112679543      1
#> 935   1.153381482      1
#> 936  -0.296407567      1
#> 937   1.384498701      1
#> 938   0.636959526      1
#> 939   2.201842645      1
#> 940   2.659747085      1
#> 941   0.818145129      1
#> 942  -0.141321312      1
#> 943  -0.285308787      1
#> 944  -0.203192590      1
#> 945   1.271144362      1
#> 946   1.410611142      1
#> 947   2.333924887      1
#> 948   1.901599110      1
#> 949  -0.743414159      1
#> 950   0.157312369      1
#> 951   0.505696324      1
#> 952   0.372441366      1
#> 953   1.875593036      1
#> 954   0.333326584      1
#> 955   0.091505358      1
#> 956   0.427895792      1
#> 957   1.234022420      1
#> 958  -1.013677977      1
#> 959   1.663795299      1
#> 960   2.351891806      1
#> 961   0.622893790      1
#> 962   0.275138218      1
#> 963   1.870613592      1
#> 964  -0.185096329      1
#> 965  -0.492797639      1
#> 966  -2.272402272      1
#> 967  -0.148697364      1
#> 968   1.016744564      1
#> 969  -0.646941766      1
#> 970   0.921898977      1
#> 971   0.537791908      1
#> 972  -0.951638351      1
#> 973   1.581931097      1
#> 974  -1.320959892      1
#> 975   0.522504335      1
#> 976   1.719799547      1
#> 977  -0.472062533      1
#> 978   0.587921292      1
#> 979   2.604160360      1
#> 980  -1.279821168      1
#> 981   1.008975524      1
#> 982   1.098800626      1
#> 983   0.831997985      1
#> 984   0.760303476      1
#> 985   0.570825809      1
#> 986   0.713573583      1
#> 987  -1.363445432      1
#> 988   0.864279348      1
#> 989   1.299089698      1
#> 990   0.312900387      1
#> 991   1.061947084      1
#> 992  -0.634726743      1
#> 993  -0.217863675      1
#> 994  -0.249969083      1
#> 995   0.682366202      1
#> 996   0.240227530      1
#> 997   0.728086504      1
#> 998   1.776644093      1
#> 999   0.891245443      1
#> 1000 -0.637327454      1