Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.314825430      0
#> 2    -0.730897650      0
#> 3    -2.525102321      0
#> 4    -0.164435118      0
#> 5     0.259863199      0
#> 6     0.563881619      0
#> 7    -1.401539938      0
#> 8    -0.492847957      0
#> 9     0.521639684      0
#> 10   -1.816694490      0
#> 11   -0.226551968      0
#> 12    0.755811291      0
#> 13   -0.562348285      0
#> 14    0.894139870      0
#> 15    0.802572703      0
#> 16    0.032536453      0
#> 17    0.972731746      0
#> 18   -0.298844401      0
#> 19   -0.295270805      0
#> 20    0.812326080      0
#> 21   -0.253811076      0
#> 22   -0.552687706      0
#> 23   -1.565736015      0
#> 24    0.184520245      0
#> 25    0.688838800      0
#> 26    0.013929270      0
#> 27   -1.480385424      0
#> 28   -0.756368271      0
#> 29    2.149373576      0
#> 30    0.504233926      0
#> 31    0.882208551      0
#> 32   -1.778381921      0
#> 33   -0.132464702      0
#> 34    1.471724131      0
#> 35   -1.231538185      0
#> 36   -0.434050077      0
#> 37    0.810141273      0
#> 38   -1.627641781      0
#> 39   -0.898800471      0
#> 40    0.527773394      0
#> 41    0.301258028      0
#> 42    2.149524658      0
#> 43    1.026249915      0
#> 44    1.042576210      0
#> 45   -1.024060364      0
#> 46    0.449722466      0
#> 47    0.911951908      0
#> 48    0.219221121      0
#> 49   -0.371637208      0
#> 50   -0.609161580      0
#> 51   -0.079709249      0
#> 52    0.471122498      0
#> 53    1.019975642      0
#> 54   -0.749755251      0
#> 55    0.410637059      0
#> 56    1.047090099      0
#> 57   -0.779732587      0
#> 58    0.887420309      0
#> 59   -0.576979884      0
#> 60   -0.298089334      0
#> 61    0.524718108      0
#> 62   -1.222172028      0
#> 63   -0.796449357      0
#> 64   -0.717398335      0
#> 65    1.554880678      0
#> 66   -0.510523605      0
#> 67    0.104387553      0
#> 68    1.379798648      0
#> 69    1.214393231      0
#> 70   -1.105101975      0
#> 71   -1.350166057      0
#> 72    1.563308879      0
#> 73   -0.842244595      0
#> 74    0.608260056      0
#> 75    0.337431477      0
#> 76    0.370963597      0
#> 77    0.875393028      0
#> 78    0.605952117      0
#> 79   -0.578508487      0
#> 80    1.441081536      0
#> 81   -1.411819127      0
#> 82    1.422064519      0
#> 83   -0.072808592      0
#> 84   -1.018596031      0
#> 85    1.027987976      0
#> 86    0.510062243      0
#> 87   -0.825316237      0
#> 88   -1.217624264      0
#> 89   -1.036005034      0
#> 90    0.424259166      0
#> 91    0.173454217      0
#> 92   -0.785487430      0
#> 93   -0.417862620      0
#> 94   -1.158887892      0
#> 95    1.110987905      0
#> 96   -1.686746465      0
#> 97    0.674213439      0
#> 98   -0.551104804      0
#> 99    1.424984242      0
#> 100   0.756207742      0
#> 101   0.342661774      0
#> 102  -1.755077838      0
#> 103   2.223084420      0
#> 104   0.232441995      0
#> 105  -0.491465119      0
#> 106   1.207926011      0
#> 107   0.343743222      0
#> 108  -0.368518592      0
#> 109  -0.317623759      0
#> 110  -1.845405075      0
#> 111  -0.433643720      0
#> 112  -0.368379426      0
#> 113  -0.053633422      0
#> 114  -0.998829408      0
#> 115  -0.030963767      0
#> 116   0.228197244      0
#> 117   1.098264700      0
#> 118   0.247378270      0
#> 119  -2.720801947      0
#> 120   0.267454642      0
#> 121  -1.466975551      0
#> 122   1.807509296      0
#> 123   1.750632658      0
#> 124   0.295061160      0
#> 125   0.348439364      0
#> 126  -1.361933284      0
#> 127  -1.193745244      0
#> 128   0.302908445      0
#> 129  -2.244007553      0
#> 130  -0.339230782      0
#> 131   0.993415625      0
#> 132  -1.346445729      0
#> 133  -0.051331839      0
#> 134  -1.178887449      0
#> 135  -1.452490517      0
#> 136  -1.375072888      0
#> 137  -0.041722841      0
#> 138   0.328033896      0
#> 139  -1.095201848      0
#> 140  -1.029932161      0
#> 141   0.325118488      0
#> 142  -0.774327392      0
#> 143  -0.226418547      0
#> 144   0.168821703      0
#> 145  -2.031324203      0
#> 146   0.105029107      0
#> 147   0.086854723      0
#> 148  -0.255690460      0
#> 149  -0.658185335      0
#> 150   0.727284003      0
#> 151   0.693266353      0
#> 152  -0.358517842      0
#> 153  -0.253769158      0
#> 154  -1.204535614      0
#> 155   0.098526636      0
#> 156  -0.779465374      0
#> 157  -0.332565356      0
#> 158  -0.995691133      0
#> 159   0.921371574      0
#> 160  -4.538099629      0
#> 161   0.466115306      0
#> 162  -0.957034446      0
#> 163   1.265139564      0
#> 164   0.695109142      0
#> 165   0.141477437      0
#> 166  -1.080952030      0
#> 167   0.102665418      0
#> 168  -0.407615718      0
#> 169  -0.287269988      0
#> 170  -0.201038906      0
#> 171  -1.327106064      0
#> 172  -0.902924557      0
#> 173   2.484802762      0
#> 174   0.219025572      0
#> 175  -0.676135049      0
#> 176   0.554032568      0
#> 177  -0.094072543      0
#> 178  -0.072123866      0
#> 179   2.070339488      0
#> 180   1.623486489      0
#> 181  -0.272128877      0
#> 182  -0.184138357      0
#> 183  -0.279324687      0
#> 184   0.094064440      0
#> 185  -1.296006205      0
#> 186   0.310363376      0
#> 187  -1.665730624      0
#> 188  -0.078708367      0
#> 189   0.479299363      0
#> 190  -0.620392686      0
#> 191  -1.828201277      0
#> 192  -0.705226488      0
#> 193  -0.406751382      0
#> 194  -0.098963655      0
#> 195  -0.606143472      0
#> 196  -0.983404934      0
#> 197   0.715577336      0
#> 198  -0.329555026      0
#> 199  -1.457570025      0
#> 200   0.133185011      0
#> 201   0.265293585      0
#> 202  -1.188873194      0
#> 203   0.148550498      0
#> 204  -1.733696725      0
#> 205   1.491612751      0
#> 206   0.104270351      0
#> 207   0.668854768      0
#> 208  -1.145235139      0
#> 209   0.778115394      0
#> 210  -0.739388688      0
#> 211   1.205431332      0
#> 212   0.036198601      0
#> 213   0.561373767      0
#> 214   0.058305677      0
#> 215  -0.326612422      0
#> 216  -0.722193292      0
#> 217  -1.384089535      0
#> 218  -1.468927485      0
#> 219  -0.750138929      0
#> 220  -0.738667764      0
#> 221   0.386166207      0
#> 222  -0.845877401      0
#> 223   0.363219315      0
#> 224   0.493862797      0
#> 225   1.992625397      0
#> 226  -1.110385171      0
#> 227   0.587556882      0
#> 228  -0.669988467      0
#> 229   1.833904878      0
#> 230  -1.838686131      0
#> 231   0.961797874      0
#> 232   0.274571008      0
#> 233   0.872891539      0
#> 234  -0.166745106      0
#> 235  -0.165768401      0
#> 236  -1.355351964      0
#> 237   1.412791946      0
#> 238   1.100321148      0
#> 239  -0.357490765      0
#> 240  -1.085510334      0
#> 241   0.969546490      0
#> 242  -0.343768823      0
#> 243  -1.569367964      0
#> 244   0.680176471      0
#> 245  -0.515480869      0
#> 246   0.536742085      0
#> 247   0.404526197      0
#> 248   0.141908872      0
#> 249  -0.048415817      0
#> 250  -0.904098785      0
#> 251   0.795610717      0
#> 252  -1.180095836      0
#> 253  -0.725953409      0
#> 254   0.453821789      0
#> 255   1.422522482      0
#> 256   0.405439235      0
#> 257  -0.150067400      0
#> 258   0.415769937      0
#> 259  -0.785366181      0
#> 260   0.472968065      0
#> 261   0.953252472      0
#> 262  -1.626044070      0
#> 263  -0.038962702      0
#> 264   0.451077695      0
#> 265   1.007474968      0
#> 266   0.083909600      0
#> 267   1.200826723      0
#> 268  -0.467398263      0
#> 269   1.496022434      0
#> 270  -0.916665175      0
#> 271  -0.178554205      0
#> 272  -2.439701197      0
#> 273   0.218843738      0
#> 274   0.645847604      0
#> 275   0.565152102      0
#> 276  -1.104508118      0
#> 277  -0.732323230      0
#> 278  -0.211457908      0
#> 279   0.776130397      0
#> 280  -0.527360702      0
#> 281  -0.765293174      0
#> 282   0.736320538      0
#> 283   0.200343550      0
#> 284  -0.860062099      0
#> 285   0.545854859      0
#> 286   0.196671009      0
#> 287  -0.340765380      0
#> 288   0.878422481      0
#> 289   1.004475098      0
#> 290  -0.842773590      0
#> 291   0.062207183      0
#> 292  -0.487273609      0
#> 293   0.734751313      0
#> 294   1.150087752      0
#> 295  -1.541651552      0
#> 296   0.679636309      0
#> 297  -0.035322023      0
#> 298   1.184762136      0
#> 299  -0.431941115      0
#> 300  -0.485222843      0
#> 301  -0.034811009      0
#> 302   1.152014603      0
#> 303   0.550632423      0
#> 304   0.608059524      0
#> 305   0.442221746      0
#> 306  -0.570973426      0
#> 307   2.475877799      0
#> 308   0.388456757      0
#> 309  -0.310561648      0
#> 310  -0.768416666      0
#> 311   1.007590196      0
#> 312   0.318540732      0
#> 313  -0.482950143      0
#> 314  -1.456207109      0
#> 315  -0.045093946      0
#> 316  -1.031072968      0
#> 317   1.536457689      0
#> 318   1.615717142      0
#> 319   0.609975675      0
#> 320  -1.317672577      0
#> 321  -1.485662866      0
#> 322  -0.495392604      0
#> 323   0.947357717      0
#> 324  -0.178206844      0
#> 325  -1.473599479      0
#> 326  -0.686602692      0
#> 327  -1.110025112      0
#> 328   0.629329394      0
#> 329  -0.638820783      0
#> 330  -0.511024083      0
#> 331  -0.536004183      0
#> 332   1.743642996      0
#> 333   0.415018881      0
#> 334  -1.089723236      0
#> 335  -0.694749787      0
#> 336  -0.513672432      0
#> 337  -1.108948521      0
#> 338  -0.549012993      0
#> 339   0.055150800      0
#> 340  -1.809401282      0
#> 341   0.713247758      0
#> 342   0.082408606      0
#> 343  -1.717424427      0
#> 344  -1.381193709      0
#> 345  -1.435274972      0
#> 346   0.001662319      0
#> 347  -1.885616400      0
#> 348   2.519629216      0
#> 349  -1.082995710      0
#> 350  -0.602148108      0
#> 351  -0.303190503      0
#> 352  -0.808718343      0
#> 353   0.208486504      0
#> 354  -0.016233963      0
#> 355  -0.182884323      0
#> 356   0.059345626      0
#> 357   0.114782729      0
#> 358   0.905770173      0
#> 359  -1.396517958      0
#> 360   0.851673068      0
#> 361   1.841600930      0
#> 362   2.524721098      0
#> 363  -0.504566262      0
#> 364   0.681312023      0
#> 365   1.863816279      0
#> 366  -0.435964192      0
#> 367   0.828169494      0
#> 368  -0.846160538      0
#> 369  -1.627347602      0
#> 370  -0.275216285      0
#> 371   0.150344613      0
#> 372  -0.297350598      0
#> 373  -0.300003983      0
#> 374  -0.643889838      0
#> 375  -0.642743304      0
#> 376   0.160415096      0
#> 377  -0.479898171      0
#> 378   0.136121130      0
#> 379  -1.217245562      0
#> 380   0.659468945      0
#> 381   0.929259446      0
#> 382  -0.969718502      0
#> 383   0.127641364      0
#> 384  -0.520260450      0
#> 385  -0.964651369      0
#> 386  -1.051362287      0
#> 387   0.180040893      0
#> 388   0.618141889      0
#> 389  -0.184176548      0
#> 390  -1.619657340      0
#> 391   1.110085971      0
#> 392  -1.834534375      0
#> 393   0.928661200      0
#> 394   0.453894459      0
#> 395  -1.386354918      0
#> 396  -1.652185187      0
#> 397  -0.451977119      0
#> 398  -1.138761400      0
#> 399   0.082964109      0
#> 400  -0.819605512      0
#> 401  -1.233642854      0
#> 402  -0.394611082      0
#> 403   0.556774004      0
#> 404   0.197246655      0
#> 405  -1.258260912      0
#> 406  -0.378748462      0
#> 407  -0.986940044      0
#> 408  -1.117361740      0
#> 409  -0.497864813      0
#> 410   1.026238434      0
#> 411   1.210299081      0
#> 412  -1.006218972      0
#> 413  -1.034003461      0
#> 414   0.065328635      0
#> 415   1.609612326      0
#> 416  -0.740949089      0
#> 417  -1.976068724      0
#> 418   0.104275471      0
#> 419  -0.365465517      0
#> 420   0.080785855      0
#> 421  -1.425012859      0
#> 422  -1.202734576      0
#> 423  -1.227665167      0
#> 424  -1.064270358      0
#> 425  -0.112204444      0
#> 426  -0.193371435      0
#> 427  -0.077184617      0
#> 428  -1.789966528      0
#> 429  -0.870297299      0
#> 430  -0.193994642      0
#> 431   1.691696766      0
#> 432   1.300771555      0
#> 433   1.321145209      0
#> 434   0.971916272      0
#> 435  -1.118670377      0
#> 436  -2.160051588      0
#> 437   0.200188481      0
#> 438  -1.071096068      0
#> 439  -0.248961212      0
#> 440  -0.054315219      0
#> 441   0.959339395      0
#> 442   1.006438304      0
#> 443   0.189466985      0
#> 444  -0.414068761      0
#> 445   1.290636765      0
#> 446  -1.423342668      0
#> 447  -1.695111159      0
#> 448   0.290973588      0
#> 449  -0.427029792      0
#> 450   0.247639299      0
#> 451   0.271219596      0
#> 452   0.555207809      0
#> 453   0.168426361      0
#> 454  -0.623065263      0
#> 455   0.223020903      0
#> 456   1.150066363      0
#> 457  -0.160884007      0
#> 458   0.843982940      0
#> 459  -0.351524751      0
#> 460  -1.707682565      0
#> 461   1.104626025      0
#> 462   1.247372502      0
#> 463  -1.022450760      0
#> 464  -1.030905006      0
#> 465  -0.943900017      0
#> 466   0.507457222      0
#> 467   0.208828584      0
#> 468   0.785696474      0
#> 469  -0.540083799      0
#> 470   0.594537646      0
#> 471  -0.562408058      0
#> 472  -1.253101825      0
#> 473  -0.870367523      0
#> 474   1.309986776      0
#> 475  -0.587624289      0
#> 476   0.060585709      0
#> 477  -0.414270774      0
#> 478   0.353977085      0
#> 479   0.202300940      0
#> 480   0.649134773      0
#> 481   0.793018934      0
#> 482  -0.051955608      0
#> 483  -0.258541372      0
#> 484  -1.103805951      0
#> 485  -1.006451238      0
#> 486  -1.736914534      0
#> 487  -1.558803936      0
#> 488  -0.006172131      0
#> 489   0.587782832      0
#> 490  -0.083091183      0
#> 491   0.024665619      0
#> 492   1.746888066      0
#> 493   0.652029983      0
#> 494   1.351825716      0
#> 495   2.435654107      0
#> 496   0.334311064      0
#> 497   1.307397120      0
#> 498  -0.449466205      0
#> 499   0.918480867      0
#> 500   1.217841658      0
#> 501   0.489135974      0
#> 502  -0.047052185      0
#> 503   1.551197652      0
#> 504  -0.480907711      0
#> 505   0.927471816      0
#> 506   0.592080548      0
#> 507  -0.836255028      0
#> 508   0.568881606      0
#> 509   0.470363727      0
#> 510   0.194402334      0
#> 511   0.807625527      0
#> 512  -0.422716354      0
#> 513  -0.496011061      0
#> 514   0.330781588      0
#> 515  -0.291906332      0
#> 516   2.074775094      0
#> 517   0.821625786      0
#> 518   0.028224980      0
#> 519  -1.333331821      0
#> 520  -0.147323534      0
#> 521  -1.100456861      0
#> 522  -1.106151851      0
#> 523   0.082776890      0
#> 524   0.528621994      0
#> 525   0.983949661      0
#> 526  -1.095542294      0
#> 527   0.594762444      0
#> 528  -0.179950151      0
#> 529  -0.034065056      0
#> 530   0.700965504      0
#> 531  -1.500211816      0
#> 532  -0.693906282      0
#> 533   0.146336360      0
#> 534   1.313612602      0
#> 535   1.475246042      0
#> 536   0.058190095      0
#> 537  -0.590456623      0
#> 538   1.535271774      0
#> 539  -0.965504048      0
#> 540   0.339396661      0
#> 541  -0.932537842      0
#> 542   1.272597997      0
#> 543   0.902436219      0
#> 544  -1.614784544      0
#> 545   0.337336931      0
#> 546  -0.534657406      0
#> 547  -1.557550228      0
#> 548  -0.309369117      0
#> 549   0.219130484      0
#> 550   0.452775657      0
#> 551   0.978157729      0
#> 552   0.221077019      0
#> 553  -1.082703795      0
#> 554   1.677210147      0
#> 555  -0.211792994      0
#> 556   1.292841585      0
#> 557  -0.397192108      0
#> 558  -0.552975800      0
#> 559  -0.479908831      0
#> 560   0.373282315      0
#> 561  -1.677896051      0
#> 562  -1.564309749      0
#> 563   0.294200788      0
#> 564   0.340087275      0
#> 565  -0.811519224      0
#> 566  -0.164935715      0
#> 567  -0.329407210      0
#> 568   0.467493590      0
#> 569   0.235308960      0
#> 570  -0.944608833      0
#> 571   1.184206830      0
#> 572   0.521724662      0
#> 573  -2.052172189      0
#> 574   0.197978110      0
#> 575  -0.715376559      0
#> 576   1.559894056      0
#> 577  -0.806516255      0
#> 578   1.843781286      0
#> 579   0.161243520      0
#> 580   0.111098478      0
#> 581  -2.684791490      0
#> 582  -0.626762812      0
#> 583  -0.528888442      0
#> 584   1.008350467      0
#> 585  -0.962499052      0
#> 586  -2.026673677      0
#> 587   0.163051103      0
#> 588   0.081337301      0
#> 589   0.275314078      0
#> 590  -0.583529110      0
#> 591   1.274912700      0
#> 592   1.868239689      0
#> 593   0.825631219      0
#> 594  -0.960443490      0
#> 595   1.589362621      0
#> 596   0.371100107      0
#> 597  -0.082921809      0
#> 598   1.142724702      0
#> 599  -0.456193266      0
#> 600   0.107733971      0
#> 601  -0.159439332      0
#> 602  -0.848671943      0
#> 603  -3.084168575      0
#> 604  -0.493735026      0
#> 605   0.099282350      0
#> 606  -0.365744706      0
#> 607   1.341317932      0
#> 608  -1.694268023      0
#> 609   1.801122818      0
#> 610   0.736453845      0
#> 611   0.120916235      0
#> 612   0.769967472      0
#> 613   0.856049354      0
#> 614   0.549967551      0
#> 615  -0.928531586      0
#> 616   0.652869500      0
#> 617  -0.164320060      0
#> 618   0.792236783      0
#> 619  -0.408833450      0
#> 620   0.158635442      0
#> 621  -0.838040585      0
#> 622   0.397201115      0
#> 623   0.300111842      0
#> 624   0.496063371      0
#> 625   0.473191138      0
#> 626  -0.974160137      0
#> 627   0.692545545      0
#> 628   0.085611393      0
#> 629   1.203936735      0
#> 630   1.646837788      0
#> 631   0.992060348      0
#> 632  -1.423971144      0
#> 633   0.907992744      0
#> 634   1.404533461      0
#> 635  -1.079502402      0
#> 636   0.308569885      0
#> 637   0.959271391      0
#> 638   1.676286922      0
#> 639   0.838708908      0
#> 640   1.889002167      0
#> 641  -0.716534677      0
#> 642  -0.364172644      0
#> 643  -0.255709690      0
#> 644  -0.229184147      0
#> 645   0.061859947      0
#> 646  -1.176849584      0
#> 647  -0.594501340      0
#> 648   0.525946462      0
#> 649  -1.788103814      0
#> 650   0.923038793      0
#> 651   1.095592351      0
#> 652  -0.043138083      0
#> 653  -0.303263650      0
#> 654  -1.342803356      0
#> 655  -0.315402710      0
#> 656  -0.429240999      0
#> 657   0.186423759      0
#> 658  -0.478684621      0
#> 659  -0.353848922      0
#> 660  -0.662671239      0
#> 661  -0.874891244      0
#> 662  -1.409754794      0
#> 663  -0.109808186      0
#> 664   1.567086528      0
#> 665  -1.983646822      0
#> 666   1.402190188      0
#> 667   0.460723298      0
#> 668  -0.011152123      0
#> 669   1.558130876      0
#> 670  -0.767422018      0
#> 671  -0.749213241      0
#> 672   1.215374226      0
#> 673   0.203955339      0
#> 674   0.664575419      0
#> 675   0.109007900      0
#> 676   2.244858233      0
#> 677   0.779644259      0
#> 678  -0.822556688      0
#> 679  -0.624471578      0
#> 680   0.603816787      0
#> 681   0.270724842      0
#> 682  -0.686569488      0
#> 683  -0.466073877      0
#> 684   0.476324180      0
#> 685  -0.530958573      0
#> 686   1.221403338      0
#> 687   1.186627035      0
#> 688  -1.479035791      0
#> 689   0.800400339      0
#> 690  -0.051022063      0
#> 691   0.198859717      0
#> 692  -0.599681570      0
#> 693  -0.354642768      0
#> 694  -0.655539766      0
#> 695   0.117880612      0
#> 696   0.673937072      0
#> 697  -1.861010870      0
#> 698  -0.207751719      0
#> 699  -1.072955746      0
#> 700   0.831817679      0
#> 701  -0.798394602      0
#> 702  -0.231096299      0
#> 703   0.509399096      0
#> 704  -0.430631641      0
#> 705  -0.205517704      0
#> 706   0.350955494      0
#> 707   0.874469881      0
#> 708   3.189829682      0
#> 709   0.379522526      0
#> 710  -1.421444314      0
#> 711  -0.351296441      0
#> 712  -0.620416534      0
#> 713   0.653088663      0
#> 714   1.116627946      0
#> 715   0.842833242      0
#> 716   0.441910290      0
#> 717   2.152287679      0
#> 718  -0.277368036      0
#> 719   0.885136063      0
#> 720   0.272206442      0
#> 721   1.422234931      0
#> 722   0.460402530      0
#> 723   1.562661673      0
#> 724  -0.202881588      0
#> 725  -0.895767242      0
#> 726  -0.510632083      0
#> 727   0.615066306      0
#> 728  -0.050544764      0
#> 729   1.120252211      0
#> 730   0.936806852      0
#> 731  -0.955781952      0
#> 732   1.842729224      0
#> 733  -0.506000048      0
#> 734  -0.268078654      0
#> 735  -0.595572439      0
#> 736  -0.824436385      0
#> 737   0.498621201      0
#> 738   1.755535419      0
#> 739   0.132621755      0
#> 740  -0.046759315      0
#> 741   1.035898286      0
#> 742  -0.675822606      0
#> 743   1.022912612      0
#> 744   1.011656402      0
#> 745   0.371382376      0
#> 746   0.027058132      0
#> 747   0.007345704      0
#> 748  -0.456673016      0
#> 749   1.204434163      0
#> 750   0.467846587      0
#> 751  -0.576220605      0
#> 752   0.008744657      0
#> 753  -1.492083098      0
#> 754   0.462430831      0
#> 755  -0.381697770      0
#> 756   0.338565866      0
#> 757   0.724886554      0
#> 758   0.279421533      0
#> 759  -0.397844480      0
#> 760   0.535069230      0
#> 761  -1.782259076      0
#> 762  -0.672920370      0
#> 763  -1.101550380      0
#> 764  -0.112905882      0
#> 765   0.908271530      0
#> 766   0.232616861      0
#> 767   0.674403368      0
#> 768   0.570319738      0
#> 769   0.195399440      0
#> 770   0.550673603      0
#> 771   1.382524223      0
#> 772  -0.898536221      0
#> 773  -1.060813796      0
#> 774   1.669287527      0
#> 775  -0.020001222      0
#> 776  -0.740608353      0
#> 777   0.192486717      0
#> 778   0.285022250      0
#> 779  -0.281698089      0
#> 780  -0.855408415      0
#> 781   0.226667438      0
#> 782  -0.063430907      0
#> 783  -0.376198240      0
#> 784   0.979328114      0
#> 785   0.134545565      0
#> 786  -1.049898774      0
#> 787   0.701250937      0
#> 788  -0.320014696      0
#> 789  -0.255428665      0
#> 790   0.212383673      0
#> 791   0.517875504      0
#> 792  -0.226544950      0
#> 793   1.441671772      0
#> 794   0.087858508      0
#> 795   0.183710957      0
#> 796  -0.676956553      0
#> 797  -0.748081742      0
#> 798  -0.920747577      0
#> 799  -0.657641600      0
#> 800  -0.624986912      0
#> 801   0.657534238      0
#> 802  -0.039910652      0
#> 803  -0.243775035      0
#> 804  -1.690349818      0
#> 805   0.255387276      0
#> 806  -0.561792973      0
#> 807  -0.017377778      0
#> 808   0.625739901      0
#> 809   0.520502934      0
#> 810  -1.445410733      0
#> 811  -1.039435726      0
#> 812   1.118570034      0
#> 813   0.146674911      0
#> 814   0.456005076      0
#> 815  -0.121266517      0
#> 816  -0.030347793      0
#> 817  -0.694621229      0
#> 818   0.271885060      0
#> 819   0.904643910      0
#> 820   0.243906478      0
#> 821   0.724747346      0
#> 822   1.917533421      0
#> 823   0.535459141      0
#> 824   0.593050726      0
#> 825   0.669565725      0
#> 826   0.185701221      0
#> 827  -1.543173721      0
#> 828  -0.469278885      0
#> 829   1.699140261      0
#> 830   0.105764412      0
#> 831  -0.765338104      0
#> 832  -0.984507728      0
#> 833   2.723331663      0
#> 834   0.423733090      0
#> 835   0.004440459      0
#> 836   1.598293724      0
#> 837   0.144870962      0
#> 838   0.034448129      0
#> 839  -0.059932368      0
#> 840  -0.518022755      0
#> 841  -1.731626859      0
#> 842   0.673563318      0
#> 843   1.582079497      0
#> 844  -0.163713385      0
#> 845   0.186787911      0
#> 846   0.367833302      0
#> 847   0.453730110      0
#> 848   1.992458759      0
#> 849   0.926930086      0
#> 850  -0.076774510      0
#> 851  -0.090347837      0
#> 852  -0.104600237      0
#> 853  -0.586238808      0
#> 854   1.809423918      0
#> 855  -0.498595166      0
#> 856   2.043834549      0
#> 857   0.256264774      0
#> 858  -0.425449305      0
#> 859   0.223730398      0
#> 860   1.923022327      0
#> 861  -0.765294425      0
#> 862  -0.954514559      0
#> 863   2.063286071      0
#> 864  -1.396874526      0
#> 865   0.507130899      0
#> 866  -1.662526158      0
#> 867   2.023240680      0
#> 868   0.221933394      0
#> 869  -0.113025580      0
#> 870   2.024267656      0
#> 871  -0.486279755      0
#> 872  -1.001944402      0
#> 873  -0.656703027      0
#> 874  -1.958593338      0
#> 875   0.058420289      0
#> 876  -1.389913501      0
#> 877  -1.305293712      0
#> 878   2.334057918      0
#> 879  -0.842969282      0
#> 880   0.364253197      0
#> 881   0.247929448      0
#> 882   0.637699055      0
#> 883   1.251817759      0
#> 884   0.215084001      0
#> 885  -0.045708091      0
#> 886  -1.654679015      0
#> 887  -0.382674063      0
#> 888   1.294725946      0
#> 889  -0.637637722      0
#> 890  -0.070884818      0
#> 891  -0.907469295      0
#> 892  -0.935277459      0
#> 893  -0.628565592      0
#> 894  -1.232405890      0
#> 895  -1.719639574      0
#> 896   2.271419033      1
#> 897   0.087044894      1
#> 898  -0.112560331      1
#> 899   0.886144861      1
#> 900  -0.128893447      1
#> 901   1.952419087      1
#> 902   1.149926342      1
#> 903   2.048811558      1
#> 904   0.062703214      1
#> 905   0.546882188      1
#> 906   1.168024713      1
#> 907   1.631572372      1
#> 908   1.048430512      1
#> 909   2.426058841      1
#> 910  -0.795607470      1
#> 911   2.724244634      1
#> 912   2.028562089      1
#> 913   1.276181980      1
#> 914   0.154482839      1
#> 915   1.013354814      1
#> 916   0.816667446      1
#> 917  -0.156917204      1
#> 918  -0.671919614      1
#> 919   0.022056248      1
#> 920  -0.115463624      1
#> 921   0.540679690      1
#> 922   0.381233195      1
#> 923   1.031313105      1
#> 924   0.505886318      1
#> 925  -0.699896108      1
#> 926   0.863811429      1
#> 927  -0.445440232      1
#> 928   0.937488351      1
#> 929   2.380251063      1
#> 930   0.994567778      1
#> 931  -0.720910358      1
#> 932   0.330004088      1
#> 933   2.254922484      1
#> 934   1.243841268      1
#> 935   0.761909583      1
#> 936  -1.969704799      1
#> 937   1.026985775      1
#> 938   1.033303413      1
#> 939   0.496579305      1
#> 940   0.860633938      1
#> 941  -0.182365965      1
#> 942   0.718787717      1
#> 943   0.343765968      1
#> 944   0.752316015      1
#> 945   0.469884184      1
#> 946   0.339978060      1
#> 947  -0.228799472      1
#> 948   0.402203656      1
#> 949   1.202177679      1
#> 950   0.245567931      1
#> 951  -0.110186558      1
#> 952  -0.262263198      1
#> 953   1.817375571      1
#> 954   0.012619275      1
#> 955   0.609628436      1
#> 956  -0.029466860      1
#> 957   1.179275846      1
#> 958   0.838525685      1
#> 959   0.461240064      1
#> 960  -0.591999083      1
#> 961  -0.176662310      1
#> 962  -0.622649148      1
#> 963   0.978618059      1
#> 964   0.730843579      1
#> 965   0.504026745      1
#> 966   0.493258826      1
#> 967   1.033489240      1
#> 968   1.368464208      1
#> 969   1.440165861      1
#> 970  -0.379474964      1
#> 971   1.531005036      1
#> 972   1.992556931      1
#> 973   0.740865523      1
#> 974   1.994480135      1
#> 975   2.119818749      1
#> 976   1.886387559      1
#> 977   0.798049618      1
#> 978   0.977486589      1
#> 979   0.299195915      1
#> 980   1.131000414      1
#> 981   0.144675985      1
#> 982   1.356594456      1
#> 983  -0.594376841      1
#> 984  -0.160268068      1
#> 985   1.489735936      1
#> 986  -0.412954617      1
#> 987   0.047555083      1
#> 988   2.342560810      1
#> 989   1.787612840      1
#> 990   2.039480932      1
#> 991   0.006764793      1
#> 992   0.823537550      1
#> 993   2.226682558      1
#> 994   0.597807922      1
#> 995   0.869755052      1
#> 996   0.014207355      1
#> 997  -0.817797007      1
#> 998  -0.849639163      1
#> 999   0.717447007      1
#> 1000  3.166989187      1