Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.351548304      0
#> 2     0.499137578      0
#> 3    -1.001344617      0
#> 4     0.684388020      0
#> 5     0.483513404      0
#> 6    -0.458038136      0
#> 7    -0.203341152      0
#> 8    -1.513815640      0
#> 9     1.549490714      0
#> 10    0.322045637      0
#> 11   -0.443895300      0
#> 12    0.645820534      0
#> 13    1.121905241      0
#> 14    1.313653383      0
#> 15    1.366213080      0
#> 16   -1.005654859      0
#> 17   -0.893689055      0
#> 18   -0.289789946      0
#> 19   -2.694890268      0
#> 20   -0.141043647      0
#> 21   -0.462083321      0
#> 22   -1.013065999      0
#> 23   -1.502660974      0
#> 24    0.053001521      0
#> 25   -0.968855463      0
#> 26   -0.962772871      0
#> 27    0.241368869      0
#> 28    0.206379994      0
#> 29    0.793545828      0
#> 30    1.913770334      0
#> 31   -0.316050009      0
#> 32    1.023529298      0
#> 33    1.107129228      0
#> 34    0.341078395      0
#> 35   -3.203716979      0
#> 36   -1.029752184      0
#> 37   -0.245135981      0
#> 38    0.161017241      0
#> 39    0.464184895      0
#> 40   -0.403210171      0
#> 41   -0.838362250      0
#> 42   -0.614423833      0
#> 43   -0.604710009      0
#> 44    1.813223810      0
#> 45    1.361977783      0
#> 46    2.293902580      0
#> 47   -0.537903083      0
#> 48   -0.032946363      0
#> 49    0.031398668      0
#> 50   -1.618763948      0
#> 51    0.900037018      0
#> 52    1.425775196      0
#> 53   -0.261057873      0
#> 54    0.363946823      0
#> 55   -0.761442436      0
#> 56    0.298346062      0
#> 57    2.263107908      0
#> 58   -1.062983850      0
#> 59    1.219399734      0
#> 60    1.688033093      0
#> 61   -0.708758876      0
#> 62    1.010322128      0
#> 63   -2.323515632      0
#> 64   -0.280739101      0
#> 65   -1.488608233      0
#> 66   -1.106319106      0
#> 67   -0.075011006      0
#> 68    1.292161981      0
#> 69   -0.452551148      0
#> 70    0.355198066      0
#> 71   -0.566920636      0
#> 72    0.308018843      0
#> 73    1.078858196      0
#> 74    0.130161615      0
#> 75    0.644378102      0
#> 76    0.542690311      0
#> 77   -0.911241645      0
#> 78    0.125104516      0
#> 79    0.019330251      0
#> 80   -1.304798952      0
#> 81    0.310822431      0
#> 82   -1.797514060      0
#> 83   -0.439031434      0
#> 84    0.522588425      0
#> 85   -0.812512276      0
#> 86   -1.830279644      0
#> 87   -0.304218124      0
#> 88    0.405938176      0
#> 89    0.558681959      0
#> 90    2.212594882      0
#> 91    0.328728383      0
#> 92    1.439732415      0
#> 93    0.088725779      0
#> 94   -0.316407336      0
#> 95   -0.573634494      0
#> 96   -0.379747060      0
#> 97   -0.242037175      0
#> 98    1.573731351      0
#> 99    0.021021926      0
#> 100   0.481780130      0
#> 101   0.149487948      0
#> 102   1.781009855      0
#> 103  -0.702763627      0
#> 104  -0.143379654      0
#> 105  -0.033455925      0
#> 106  -0.964603544      0
#> 107   0.151018933      0
#> 108  -0.441659915      0
#> 109  -0.050036465      0
#> 110   0.258312990      0
#> 111  -1.310691595      0
#> 112   0.820190479      0
#> 113  -1.725368180      0
#> 114   0.438639699      0
#> 115  -0.779464834      0
#> 116   0.828653645      0
#> 117   0.733698599      0
#> 118   0.138549915      0
#> 119  -0.904145104      0
#> 120   0.430566426      0
#> 121  -0.377565889      0
#> 122  -0.656608706      0
#> 123  -0.555868732      0
#> 124  -0.264153147      0
#> 125  -0.231612812      0
#> 126   1.851331600      0
#> 127  -0.437986720      0
#> 128  -1.235659171      0
#> 129  -1.080531530      0
#> 130   1.491499539      0
#> 131   0.520597557      0
#> 132  -0.364658307      0
#> 133   0.758184985      0
#> 134  -1.475758776      0
#> 135   0.020029531      0
#> 136  -0.875032869      0
#> 137  -1.477310351      0
#> 138  -0.881926555      0
#> 139  -0.393945684      0
#> 140   0.468072166      0
#> 141   0.722363617      0
#> 142   1.644750004      0
#> 143  -0.227893864      0
#> 144   0.976919345      0
#> 145  -1.627071385      0
#> 146  -0.884626092      0
#> 147  -2.123271515      0
#> 148   0.346757721      0
#> 149  -1.609438200      0
#> 150   0.583030698      0
#> 151   0.808145073      0
#> 152   0.625548867      0
#> 153   0.442661655      0
#> 154  -1.266684394      0
#> 155   1.806415677      0
#> 156   0.944176861      0
#> 157   2.533148410      0
#> 158   0.420126854      0
#> 159  -0.199622075      0
#> 160  -0.381588967      0
#> 161  -0.259039897      0
#> 162  -0.400428033      0
#> 163  -0.436311332      0
#> 164   0.896682722      0
#> 165   0.415971740      0
#> 166   1.393033054      0
#> 167   1.146527395      0
#> 168  -1.384380932      0
#> 169   1.221354788      0
#> 170  -1.084844614      0
#> 171  -0.026841852      0
#> 172  -0.520971769      0
#> 173  -0.261053114      0
#> 174   0.941040350      0
#> 175  -1.772671414      0
#> 176   0.941522836      0
#> 177   0.638866928      0
#> 178  -0.334641922      0
#> 179  -0.314094813      0
#> 180   0.064309377      0
#> 181  -0.337803814      0
#> 182  -1.035016071      0
#> 183   0.084623058      0
#> 184   0.071438070      0
#> 185   1.283751991      0
#> 186  -2.155676695      0
#> 187   0.398246519      0
#> 188  -0.079267671      0
#> 189   0.374899750      0
#> 190  -1.569861008      0
#> 191  -0.392977703      0
#> 192   0.422619351      0
#> 193   1.971948608      0
#> 194  -0.377378452      0
#> 195   0.401663764      0
#> 196   0.866903356      0
#> 197  -0.987495726      0
#> 198  -0.968494487      0
#> 199   0.747578354      0
#> 200  -0.694759823      0
#> 201   1.002055173      0
#> 202  -0.741978745      0
#> 203  -0.757609818      0
#> 204  -0.797254357      0
#> 205   0.590590348      0
#> 206  -1.632171519      0
#> 207   1.457816693      0
#> 208   0.261850269      0
#> 209  -1.093796018      0
#> 210  -1.490910901      0
#> 211  -0.241130154      0
#> 212   0.453731785      0
#> 213  -0.451148754      0
#> 214   0.173761540      0
#> 215   2.083863858      0
#> 216   0.366419219      0
#> 217   1.660942319      0
#> 218   1.253359632      0
#> 219   0.124356918      0
#> 220   0.635973470      0
#> 221   0.362565023      0
#> 222   1.530098339      0
#> 223  -1.288475379      0
#> 224   0.950819547      0
#> 225   0.132826317      0
#> 226   0.216931613      0
#> 227   0.014051370      0
#> 228   2.318963018      0
#> 229  -0.739559318      0
#> 230  -1.308096534      0
#> 231   0.970982056      0
#> 232  -0.223856294      0
#> 233   1.224517402      0
#> 234   1.558185158      0
#> 235  -1.857271687      0
#> 236  -1.053848595      0
#> 237  -1.857768327      0
#> 238   0.109581006      0
#> 239  -1.300005470      0
#> 240  -0.715324579      0
#> 241  -0.526393594      0
#> 242   0.949771297      0
#> 243   0.017769264      0
#> 244   0.989588963      0
#> 245   0.896086179      0
#> 246  -0.072308290      0
#> 247  -1.124950994      0
#> 248  -0.043307378      0
#> 249   0.355689357      0
#> 250   0.072713652      0
#> 251   0.850540166      0
#> 252   0.251427388      0
#> 253   2.110096157      0
#> 254  -1.206710250      0
#> 255  -0.748627554      0
#> 256   1.189267532      0
#> 257   0.240420211      0
#> 258  -0.756186340      0
#> 259   0.106706012      0
#> 260  -0.105982148      0
#> 261   0.279399379      0
#> 262  -0.040887008      0
#> 263  -0.601165723      0
#> 264  -0.097878867      0
#> 265   1.135105299      0
#> 266   0.877390691      0
#> 267   1.696569209      0
#> 268  -2.075593154      0
#> 269   1.245214643      0
#> 270   0.695619497      0
#> 271  -1.051424050      0
#> 272   1.767543504      0
#> 273  -0.069576955      0
#> 274   1.580705378      0
#> 275  -1.103193920      0
#> 276   0.250026085      0
#> 277   1.493400405      0
#> 278   0.237833625      0
#> 279  -0.382789922      0
#> 280  -0.432717100      0
#> 281   1.837447487      0
#> 282   0.995804940      0
#> 283  -1.729752499      0
#> 284  -0.046390493      0
#> 285  -0.432825096      0
#> 286   0.557178167      0
#> 287   0.435803048      0
#> 288  -0.143396656      0
#> 289  -0.590036673      0
#> 290   0.012727652      0
#> 291  -0.833123006      0
#> 292  -0.861834094      0
#> 293  -1.012642881      0
#> 294  -0.041692057      0
#> 295  -0.833963699      0
#> 296  -1.790483326      0
#> 297   0.066027908      0
#> 298  -0.296807034      0
#> 299  -0.932556561      0
#> 300   0.729421087      0
#> 301   0.023967024      0
#> 302  -0.493050028      0
#> 303   0.470506490      0
#> 304  -0.687737252      0
#> 305   0.530814636      0
#> 306  -1.759330490      0
#> 307  -0.473181124      0
#> 308   0.251586374      0
#> 309  -0.229416655      0
#> 310   0.723287202      0
#> 311  -0.863966998      0
#> 312  -0.412338473      0
#> 313   0.889894267      0
#> 314   1.066555505      0
#> 315   0.633979475      0
#> 316  -0.671157690      0
#> 317  -0.080281776      0
#> 318   0.197114965      0
#> 319   1.254828393      0
#> 320  -0.743360840      0
#> 321   1.277254295      0
#> 322  -1.076979900      0
#> 323  -0.872788889      0
#> 324  -0.651918065      0
#> 325  -1.589726989      0
#> 326   1.577835831      0
#> 327   1.133735129      0
#> 328  -0.001449185      0
#> 329   0.563441476      0
#> 330   1.923430695      0
#> 331  -0.374397045      0
#> 332  -0.159479240      0
#> 333   1.095689237      0
#> 334  -0.384534317      0
#> 335  -0.415404938      0
#> 336   0.205606129      0
#> 337   1.044422017      0
#> 338  -1.171027303      0
#> 339  -0.255035010      0
#> 340   0.820381878      0
#> 341  -1.522838334      0
#> 342  -0.726942159      0
#> 343   0.113568748      0
#> 344  -1.757470485      0
#> 345  -0.180369841      0
#> 346   1.409690329      0
#> 347   0.442703055      0
#> 348  -1.134864303      0
#> 349   0.502359243      0
#> 350   0.954357273      0
#> 351  -0.573594534      0
#> 352   0.654004611      0
#> 353  -0.422141317      0
#> 354   0.874929313      0
#> 355  -0.339186352      0
#> 356  -1.532577372      0
#> 357  -0.953722257      0
#> 358   0.442624802      0
#> 359   0.364261891      0
#> 360  -0.301867738      0
#> 361  -0.295021020      0
#> 362  -1.466949669      0
#> 363   0.555958955      0
#> 364  -1.416360297      0
#> 365   0.031214706      0
#> 366   1.621769135      0
#> 367   1.177970440      0
#> 368  -0.931865608      0
#> 369   0.372977156      0
#> 370   0.738184611      0
#> 371  -0.350612230      0
#> 372   0.135792882      0
#> 373   0.456151458      0
#> 374  -0.412337444      0
#> 375  -0.748908135      0
#> 376  -0.385818896      0
#> 377   1.040202254      0
#> 378  -2.642360453      0
#> 379   1.745142064      0
#> 380   0.378659770      0
#> 381   0.109471089      0
#> 382  -0.583032991      0
#> 383  -0.307953152      0
#> 384   1.143317219      0
#> 385   1.243602122      0
#> 386  -1.085817921      0
#> 387  -1.085260904      0
#> 388   0.994150867      0
#> 389  -0.201818196      0
#> 390  -0.613658099      0
#> 391   0.388246599      0
#> 392  -0.687647887      0
#> 393  -0.985324365      0
#> 394   1.089789811      0
#> 395  -2.367878232      0
#> 396   2.182806503      0
#> 397  -1.147060114      0
#> 398   0.413542459      0
#> 399  -0.125601520      0
#> 400  -1.524437884      0
#> 401  -0.896364875      0
#> 402   0.441396278      0
#> 403  -2.015869720      0
#> 404  -0.994430208      0
#> 405  -0.632014914      0
#> 406   0.632614285      0
#> 407   1.735510789      0
#> 408   0.404767711      0
#> 409   2.082948311      0
#> 410  -0.434964698      0
#> 411  -0.052540453      0
#> 412  -0.291626694      0
#> 413  -0.601883336      0
#> 414   0.197729410      0
#> 415  -0.712820989      0
#> 416   0.734665430      0
#> 417   0.221629638      0
#> 418  -0.276786690      0
#> 419   1.329280631      0
#> 420  -1.835323308      0
#> 421  -0.882750456      0
#> 422   0.556468936      0
#> 423  -0.784006194      0
#> 424  -0.346772982      0
#> 425  -1.503025792      0
#> 426   0.689552674      0
#> 427  -0.455368655      0
#> 428   0.475009058      0
#> 429   0.557097030      0
#> 430  -0.840088306      0
#> 431  -0.263312021      0
#> 432  -2.382817159      0
#> 433   0.791681614      0
#> 434  -1.139584428      0
#> 435  -0.764956086      0
#> 436   0.367967178      0
#> 437   0.333543569      0
#> 438   0.797068258      0
#> 439  -0.317418360      0
#> 440   0.205721209      0
#> 441  -1.827045029      0
#> 442  -0.602477867      0
#> 443  -1.306642630      0
#> 444   1.744737277      0
#> 445   1.438553437      0
#> 446  -0.629508264      0
#> 447   0.170049934      0
#> 448   0.188786989      0
#> 449  -3.396285318      0
#> 450  -0.761327065      0
#> 451   0.524667967      0
#> 452  -1.001143275      0
#> 453  -0.162436711      0
#> 454  -0.417740840      0
#> 455  -0.305887798      0
#> 456  -0.662952687      0
#> 457  -1.987497183      0
#> 458   0.020508772      0
#> 459  -0.781233568      0
#> 460   1.016790265      0
#> 461  -1.049672362      0
#> 462  -0.743547484      0
#> 463  -0.478927471      0
#> 464   0.332602516      0
#> 465   0.355765036      0
#> 466   1.252880400      0
#> 467   0.934065653      0
#> 468   0.665545364      0
#> 469  -0.731031676      0
#> 470  -0.492613542      0
#> 471  -0.562050847      0
#> 472   0.740282895      0
#> 473  -0.044726288      0
#> 474   0.530760637      0
#> 475   1.510672440      0
#> 476  -0.223260818      0
#> 477   0.422589629      0
#> 478  -0.354863144      0
#> 479  -0.984428150      0
#> 480   0.729206631      0
#> 481  -0.701988293      0
#> 482  -0.187848874      0
#> 483  -0.145827219      0
#> 484  -0.375030516      0
#> 485   0.027810033      0
#> 486  -0.368917627      0
#> 487  -2.281566633      0
#> 488  -1.115110966      0
#> 489  -0.777184679      0
#> 490  -0.858488047      0
#> 491  -0.151184965      0
#> 492  -0.160887814      0
#> 493   1.738443797      0
#> 494   2.665609308      0
#> 495  -0.178218413      0
#> 496   0.493742995      0
#> 497  -2.840854950      0
#> 498   0.175056610      0
#> 499   2.348137377      0
#> 500  -1.238232315      0
#> 501  -0.033002184      0
#> 502   0.495240678      0
#> 503  -1.266791953      0
#> 504   0.556526206      0
#> 505  -1.095894561      0
#> 506  -0.579816434      0
#> 507  -0.395180160      0
#> 508   1.329674264      0
#> 509   0.097283140      0
#> 510  -0.678577980      0
#> 511   1.107341467      0
#> 512  -1.301620774      0
#> 513  -1.221716850      0
#> 514   1.166486029      0
#> 515   0.260283717      0
#> 516   0.268410255      0
#> 517   0.453173369      0
#> 518  -0.023396220      0
#> 519  -0.026675122      0
#> 520   0.027398455      0
#> 521   0.918122633      0
#> 522  -1.427088354      0
#> 523   1.148419197      0
#> 524  -0.340271885      0
#> 525   0.040749267      0
#> 526   0.521216207      0
#> 527  -0.731040628      0
#> 528   0.034039051      0
#> 529  -0.152469753      0
#> 530   0.445453517      0
#> 531  -0.114350960      0
#> 532  -0.688747469      0
#> 533   0.174617815      0
#> 534   0.067359270      0
#> 535  -0.424830033      0
#> 536  -1.037234691      0
#> 537   1.089264052      0
#> 538  -1.183955198      0
#> 539   0.970031809      0
#> 540  -0.285263791      0
#> 541  -0.302551269      0
#> 542  -1.747176234      0
#> 543   2.196904380      0
#> 544   1.129180560      0
#> 545  -1.724459519      0
#> 546  -0.981668910      0
#> 547  -1.407627907      0
#> 548  -1.938468272      0
#> 549   1.182488076      0
#> 550  -0.467446635      0
#> 551   0.717247760      0
#> 552  -0.119914983      0
#> 553  -1.148727039      0
#> 554   0.158607876      0
#> 555  -1.228309938      0
#> 556  -1.077223585      0
#> 557   0.725976997      0
#> 558  -0.644010182      0
#> 559  -1.101443772      0
#> 560  -1.105555848      0
#> 561   1.892094031      0
#> 562   0.284556479      0
#> 563   0.808777581      0
#> 564   1.842848421      0
#> 565  -0.284901513      0
#> 566   0.955073830      0
#> 567   0.955288439      0
#> 568   1.141828065      0
#> 569   1.081993467      0
#> 570   0.493859868      0
#> 571  -1.331127537      0
#> 572   0.003542154      0
#> 573  -0.106043330      0
#> 574  -1.708091274      0
#> 575  -0.862812253      0
#> 576   2.125724523      0
#> 577   0.291797932      0
#> 578   0.199856226      0
#> 579  -1.609349989      0
#> 580   0.555648167      0
#> 581   0.778857739      0
#> 582  -0.310062888      0
#> 583   0.394476388      0
#> 584  -1.336782744      0
#> 585  -1.012134804      0
#> 586  -0.173444970      0
#> 587  -0.026658957      0
#> 588  -0.187714261      0
#> 589   0.554749769      0
#> 590  -0.027605200      0
#> 591  -0.479510527      0
#> 592   1.599338565      0
#> 593   0.719187394      0
#> 594  -1.477299342      0
#> 595   0.885524078      0
#> 596  -0.279608023      0
#> 597   0.325255470      0
#> 598  -0.198689356      0
#> 599  -0.646895737      0
#> 600  -0.897275878      0
#> 601   2.091788987      0
#> 602  -1.086617864      0
#> 603  -0.634394152      0
#> 604  -1.160200593      0
#> 605   0.912689546      0
#> 606   0.669794695      0
#> 607  -0.742696760      0
#> 608  -0.586092020      0
#> 609   2.190408537      0
#> 610  -1.484323410      0
#> 611  -0.176520941      0
#> 612  -1.480280885      0
#> 613  -1.226392800      0
#> 614  -1.591493738      0
#> 615   1.332767335      0
#> 616   0.253276861      0
#> 617  -0.909278600      0
#> 618   0.602248212      0
#> 619  -0.454645236      0
#> 620  -0.611266573      0
#> 621  -1.044597016      0
#> 622  -1.505929361      0
#> 623   0.292541676      0
#> 624  -1.403485810      0
#> 625   1.475427869      0
#> 626  -1.177600824      0
#> 627  -1.648193483      0
#> 628  -1.031499187      0
#> 629   0.345723487      0
#> 630   0.330404112      0
#> 631   0.293809787      0
#> 632  -0.315226106      0
#> 633  -0.485847911      0
#> 634   0.865298420      0
#> 635  -0.147972194      0
#> 636   1.942943194      0
#> 637   1.222261352      0
#> 638  -0.981874539      0
#> 639   0.773063030      0
#> 640   1.766795849      0
#> 641   1.080868864      0
#> 642   0.557069779      0
#> 643  -0.665705279      0
#> 644  -0.391640731      0
#> 645   0.918825745      0
#> 646   1.083502734      0
#> 647   0.586534010      0
#> 648   0.620654339      0
#> 649   1.336078371      0
#> 650  -0.143208070      0
#> 651  -0.994931859      0
#> 652   0.124743284      0
#> 653  -0.245768217      0
#> 654  -0.263912079      0
#> 655  -1.099636739      0
#> 656   0.605708709      0
#> 657  -0.707956815      0
#> 658   0.316980834      0
#> 659   2.092145368      0
#> 660  -0.347245421      0
#> 661  -0.653614037      0
#> 662   0.115392403      0
#> 663   0.730828176      0
#> 664  -0.098925041      0
#> 665  -0.441498546      0
#> 666   0.702219237      0
#> 667   0.014917475      0
#> 668  -1.634883771      0
#> 669   0.787886190      0
#> 670   0.037285122      0
#> 671  -0.173198257      0
#> 672  -2.269997665      0
#> 673   0.917926131      0
#> 674   0.909326253      0
#> 675   1.169487332      0
#> 676  -0.703012918      0
#> 677   0.072786251      0
#> 678   0.148112065      0
#> 679   1.402539450      0
#> 680   0.258795239      0
#> 681  -0.664480088      0
#> 682  -0.576691499      0
#> 683  -1.585280901      0
#> 684   0.432449884      0
#> 685   1.777252400      0
#> 686   0.262068292      0
#> 687   0.451256019      0
#> 688   0.848134430      0
#> 689  -1.189424597      0
#> 690   1.264301410      0
#> 691  -1.119385505      0
#> 692  -0.323543031      0
#> 693  -0.939445158      0
#> 694  -0.945310149      0
#> 695   0.702196561      0
#> 696   0.691971699      0
#> 697   2.998546317      0
#> 698  -1.741829827      0
#> 699  -0.248317020      0
#> 700  -1.498478867      0
#> 701  -0.068619316      0
#> 702  -0.116145598      0
#> 703   0.641660434      0
#> 704   1.134013015      0
#> 705  -1.134534381      0
#> 706   0.154549404      0
#> 707  -1.340694434      0
#> 708  -0.027510869      0
#> 709  -0.955521135      0
#> 710  -0.061962234      0
#> 711   0.788242292      0
#> 712  -1.396957208      0
#> 713  -0.195960289      0
#> 714   0.025578622      0
#> 715   0.332749926      0
#> 716  -1.016329501      0
#> 717   0.356634359      0
#> 718  -0.186875089      0
#> 719  -0.665314909      0
#> 720   1.072413529      0
#> 721  -0.563590108      0
#> 722  -0.871007303      0
#> 723  -1.063407796      0
#> 724   0.103540230      0
#> 725  -1.031245367      0
#> 726  -1.622643243      0
#> 727  -0.136845410      0
#> 728   0.929863719      0
#> 729   0.126180864      0
#> 730  -1.660113978      0
#> 731   0.701345895      0
#> 732   0.731602334      0
#> 733   1.109538632      0
#> 734  -1.771620140      0
#> 735   0.431274224      0
#> 736  -1.166518780      0
#> 737   0.173954883      0
#> 738  -0.214873025      0
#> 739  -0.563423499      0
#> 740  -0.733385559      0
#> 741   1.731065380      0
#> 742  -0.652720836      0
#> 743   1.776864448      0
#> 744   0.715246179      0
#> 745  -0.340893001      0
#> 746  -1.241800271      0
#> 747  -0.475532091      0
#> 748   0.350087960      0
#> 749  -0.088294060      0
#> 750  -0.969268985      0
#> 751  -0.892496913      0
#> 752  -0.780327916      0
#> 753  -0.069432834      0
#> 754  -0.534211895      0
#> 755  -0.735062173      0
#> 756  -0.099767265      0
#> 757   1.280289272      0
#> 758   0.203276999      0
#> 759   0.271616178      0
#> 760   0.444086118      0
#> 761   0.407389172      0
#> 762   0.560822880      0
#> 763  -0.372341471      0
#> 764   0.332724704      0
#> 765   0.146134440      0
#> 766   1.790794459      0
#> 767  -0.924196013      0
#> 768  -1.892894565      0
#> 769  -1.425501163      0
#> 770   0.087368497      0
#> 771  -2.182359069      0
#> 772   1.316146282      0
#> 773   1.550723689      0
#> 774  -0.134979442      0
#> 775   0.191068039      0
#> 776   2.035440326      0
#> 777   0.524341276      0
#> 778  -1.989770021      0
#> 779   2.013861004      0
#> 780  -0.502622068      0
#> 781   0.271829792      0
#> 782  -0.392756275      0
#> 783   0.654025646      0
#> 784  -0.875554595      0
#> 785  -0.425929691      0
#> 786   0.866891332      0
#> 787   0.270887944      0
#> 788  -0.412333628      0
#> 789  -2.291063357      0
#> 790  -0.403710207      0
#> 791  -0.187353459      0
#> 792  -0.650665750      0
#> 793  -0.983676911      0
#> 794   0.287242090      0
#> 795  -1.424974587      0
#> 796  -0.868592803      0
#> 797   1.679404988      0
#> 798  -0.730248974      0
#> 799  -1.134093394      0
#> 800  -0.297916841      0
#> 801  -0.513046466      0
#> 802  -0.213454566      0
#> 803   0.039604874      0
#> 804   0.971386254      0
#> 805   0.641220081      0
#> 806  -0.237678537      0
#> 807  -0.853712832      0
#> 808  -0.832197515      0
#> 809  -0.327055845      0
#> 810  -2.046084026      0
#> 811   0.874316729      0
#> 812   1.315607984      0
#> 813  -0.723366378      0
#> 814   1.839587004      0
#> 815   0.989809309      0
#> 816  -1.069728091      0
#> 817   0.711500067      0
#> 818   0.157764190      0
#> 819  -1.051977616      0
#> 820  -1.509814392      0
#> 821  -1.272955085      0
#> 822  -0.293083936      0
#> 823   0.600171929      0
#> 824   0.765564573      0
#> 825   0.456467178      0
#> 826   0.509528778      0
#> 827  -0.763997894      0
#> 828   0.769732109      0
#> 829  -0.078098391      0
#> 830  -1.188358607      0
#> 831   0.740588246      0
#> 832  -0.095132970      0
#> 833   1.231035196      0
#> 834   0.126555042      0
#> 835   0.429317808      0
#> 836  -0.117390435      0
#> 837   0.734334647      0
#> 838  -0.106817013      0
#> 839   1.071162866      0
#> 840  -0.569295394      0
#> 841   0.726591060      0
#> 842  -1.075364206      0
#> 843   0.772363511      0
#> 844  -0.531799248      0
#> 845  -0.540272543      0
#> 846   0.362683345      0
#> 847  -0.789130174      0
#> 848  -0.812923775      0
#> 849   1.671745360      0
#> 850   0.715039157      0
#> 851  -1.795768775      0
#> 852  -1.022394016      0
#> 853   0.701563172      0
#> 854  -0.191457653      0
#> 855   1.698308594      0
#> 856   1.487142887      0
#> 857  -0.379711672      0
#> 858   0.245791553      0
#> 859   0.168641846      0
#> 860  -0.143589328      0
#> 861  -0.635851352      0
#> 862  -0.458835489      0
#> 863  -0.583821547      0
#> 864   1.230824638      0
#> 865  -0.545439561      0
#> 866  -1.711079100      0
#> 867  -0.486558169      0
#> 868  -1.332437438      0
#> 869   1.192837072      0
#> 870   0.393170343      0
#> 871   1.109817917      0
#> 872  -0.135597831      0
#> 873   1.611446751      0
#> 874  -1.354527818      0
#> 875   0.141535568      0
#> 876   0.148549584      0
#> 877   0.083120385      0
#> 878   0.892794290      0
#> 879  -1.574226943      0
#> 880   1.276005683      0
#> 881  -0.718657877      0
#> 882  -0.817725948      0
#> 883  -0.964581335      0
#> 884   0.991617345      0
#> 885  -0.116994141      0
#> 886   0.697976191      0
#> 887   1.001434769      0
#> 888   1.286037285      0
#> 889  -1.084759806      0
#> 890   0.580362209      0
#> 891  -0.331901611      0
#> 892   0.212258109      0
#> 893  -0.142947830      0
#> 894  -0.441164502      0
#> 895   0.883553135      0
#> 896  -0.336444779      0
#> 897   0.361314650      0
#> 898  -0.073786980      0
#> 899   2.829784415      0
#> 900   0.824492727      0
#> 901  -0.660388902      0
#> 902   0.336205935      0
#> 903   0.183817058      0
#> 904  -1.323958414      0
#> 905   0.343809283      0
#> 906   1.549881618      0
#> 907   0.610588282      0
#> 908   0.053872573      0
#> 909  -0.831086564      0
#> 910   0.883862559      0
#> 911   0.665507631      0
#> 912   1.617950551      1
#> 913   1.365324244      1
#> 914  -0.370807326      1
#> 915   2.030569413      1
#> 916   1.983576840      1
#> 917   0.479148167      1
#> 918   0.388797948      1
#> 919   0.382425754      1
#> 920  -0.113470623      1
#> 921  -0.149631161      1
#> 922   0.732407764      1
#> 923   0.670747757      1
#> 924   2.047160983      1
#> 925   1.204047403      1
#> 926   0.198091532      1
#> 927  -0.112696566      1
#> 928   0.662953325      1
#> 929  -0.395016356      1
#> 930   2.500384694      1
#> 931  -0.214487443      1
#> 932   1.364978960      1
#> 933   0.935882029      1
#> 934   0.440358666      1
#> 935   1.273151077      1
#> 936   2.219696386      1
#> 937   1.455964672      1
#> 938   0.894338678      1
#> 939   1.477048362      1
#> 940   1.042050182      1
#> 941  -1.104863031      1
#> 942  -1.243873129      1
#> 943   1.119605229      1
#> 944  -1.549238569      1
#> 945   1.745683058      1
#> 946   1.798957172      1
#> 947   1.240151370      1
#> 948   0.413260485      1
#> 949   0.576035791      1
#> 950  -0.865911105      1
#> 951  -0.108211816      1
#> 952  -0.299099757      1
#> 953  -0.394614040      1
#> 954   1.378150953      1
#> 955   2.548417942      1
#> 956   0.578610897      1
#> 957   1.335203890      1
#> 958   1.369918692      1
#> 959   0.372802431      1
#> 960   1.036675824      1
#> 961  -0.319615522      1
#> 962   0.111707233      1
#> 963   1.570018906      1
#> 964  -2.539878333      1
#> 965   0.703306296      1
#> 966  -0.927155570      1
#> 967   2.401388535      1
#> 968   1.225255629      1
#> 969   0.227769611      1
#> 970   2.019383949      1
#> 971   0.565426739      1
#> 972   0.512714150      1
#> 973  -0.216381928      1
#> 974   1.397816660      1
#> 975   0.894968229      1
#> 976   1.847929356      1
#> 977   2.139244406      1
#> 978  -0.641177536      1
#> 979   0.253132061      1
#> 980  -0.042897783      1
#> 981   0.880770134      1
#> 982  -0.123731121      1
#> 983   0.345114304      1
#> 984   2.063586127      1
#> 985   2.395698858      1
#> 986   0.353336245      1
#> 987  -0.048922356      1
#> 988   0.634777478      1
#> 989  -1.929048528      1
#> 990   1.102389442      1
#> 991   0.437857118      1
#> 992  -0.411677393      1
#> 993   0.817952965      1
#> 994  -0.822411155      1
#> 995  -0.427433530      1
#> 996  -0.597505649      1
#> 997   1.804355812      1
#> 998   0.080800955      1
#> 999   1.516575565      1
#> 1000  1.075281524      1