Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.582105799      0
#> 2     0.576645884      0
#> 3    -0.857068531      0
#> 4    -0.013530887      0
#> 5    -0.388811567      0
#> 6    -0.394152366      0
#> 7    -1.109930838      0
#> 8     0.676509845      0
#> 9     2.065370601      0
#> 10    1.323999202      0
#> 11    0.957731285      0
#> 12   -0.032653365      0
#> 13   -1.669687811      0
#> 14    3.309493994      0
#> 15   -2.268175874      0
#> 16    0.447492936      0
#> 17    0.395128133      0
#> 18   -1.052919881      0
#> 19   -0.005056017      0
#> 20   -1.869376627      0
#> 21    0.712283439      0
#> 22   -0.458620571      0
#> 23   -1.090365340      0
#> 24   -0.427529640      0
#> 25    0.750139381      0
#> 26   -0.252645769      0
#> 27    0.864372704      0
#> 28    1.035390323      0
#> 29   -0.820471922      0
#> 30    2.889737990      0
#> 31   -0.277341994      0
#> 32    0.907624192      0
#> 33    1.001620520      0
#> 34    0.697040563      0
#> 35    0.889653157      0
#> 36    0.705584928      0
#> 37   -1.787342154      0
#> 38   -1.124779364      0
#> 39    0.908103974      0
#> 40    0.015074892      0
#> 41   -0.904240598      0
#> 42    0.502709381      0
#> 43   -0.211703643      0
#> 44    0.658884566      0
#> 45   -0.088596620      0
#> 46   -1.668235745      0
#> 47    1.378038321      0
#> 48   -0.244986362      0
#> 49   -0.344881150      0
#> 50    0.021134939      0
#> 51   -0.358359784      0
#> 52   -0.120486065      0
#> 53   -0.401732012      0
#> 54    1.268371283      0
#> 55   -1.942039873      0
#> 56    0.121983313      0
#> 57   -0.396006527      0
#> 58    0.729353573      0
#> 59    0.678436488      0
#> 60    0.928485716      0
#> 61    0.474679888      0
#> 62   -1.320716638      0
#> 63   -0.659854789      0
#> 64    1.021005114      0
#> 65   -0.911963305      0
#> 66    0.263595235      0
#> 67    1.797401570      0
#> 68    0.242886914      0
#> 69    0.097394083      0
#> 70   -0.288062015      0
#> 71   -0.871388493      0
#> 72   -0.054356783      0
#> 73    0.847504316      0
#> 74   -1.118844451      0
#> 75    1.975536897      0
#> 76    0.911641855      0
#> 77    0.624236505      0
#> 78   -0.159384024      0
#> 79    2.064690113      0
#> 80    0.373453660      0
#> 81    0.132010449      0
#> 82    0.717395825      0
#> 83    0.296048259      0
#> 84   -2.101186479      0
#> 85   -1.017039325      0
#> 86   -0.158616522      0
#> 87    0.582385762      0
#> 88    2.596629345      0
#> 89   -1.342871982      0
#> 90    0.668426082      0
#> 91   -0.733372885      0
#> 92   -0.840959509      0
#> 93    0.833766955      0
#> 94    0.697159418      0
#> 95   -0.764203618      0
#> 96    0.326451135      0
#> 97   -1.174714007      0
#> 98   -0.355441434      0
#> 99   -0.490972617      0
#> 100  -0.155190832      0
#> 101   0.799352983      0
#> 102  -0.484659882      0
#> 103  -0.278201062      0
#> 104  -0.489429992      0
#> 105   1.250572152      0
#> 106  -0.265494205      0
#> 107  -0.916799585      0
#> 108  -0.412178167      0
#> 109  -0.474794937      0
#> 110   0.239960066      0
#> 111  -0.533328765      0
#> 112   0.541570825      0
#> 113   0.960706726      0
#> 114   0.386008129      0
#> 115   0.363589782      0
#> 116   0.307233609      0
#> 117   1.148305446      0
#> 118   0.548019365      0
#> 119   0.387701884      0
#> 120   0.295951773      0
#> 121  -0.418692665      0
#> 122  -1.023598494      0
#> 123   0.974691619      0
#> 124   0.503407648      0
#> 125   1.217052986      0
#> 126   0.412704636      0
#> 127   0.847965935      0
#> 128   0.426575219      0
#> 129   0.385198382      0
#> 130   0.470728780      0
#> 131  -0.730225393      0
#> 132   0.897461079      0
#> 133   0.439540221      0
#> 134  -0.458564939      0
#> 135   1.217788068      0
#> 136  -0.979936084      0
#> 137  -0.412886902      0
#> 138   0.697512400      0
#> 139   0.375071678      0
#> 140   0.009045420      0
#> 141  -1.822385597      0
#> 142   1.852800220      0
#> 143  -0.923202605      0
#> 144   0.032953384      0
#> 145  -0.134534707      0
#> 146   1.897907253      0
#> 147  -0.988518551      0
#> 148   0.130724479      0
#> 149  -0.257387556      0
#> 150  -2.131957706      0
#> 151   1.299269438      0
#> 152   0.041953081      0
#> 153   0.091857727      0
#> 154  -1.795760280      0
#> 155   0.095276250      0
#> 156   0.252851205      0
#> 157   0.684756664      0
#> 158   0.883313937      0
#> 159   0.328431157      0
#> 160   0.670889232      0
#> 161   0.518001139      0
#> 162  -2.143316297      0
#> 163   0.356947974      0
#> 164  -0.327219351      0
#> 165  -0.216328322      0
#> 166  -0.063351113      0
#> 167  -0.635806615      0
#> 168   0.936587970      0
#> 169   0.743772406      0
#> 170   0.642575236      0
#> 171   0.118447973      0
#> 172   0.107409788      0
#> 173   1.794725865      0
#> 174  -0.736369922      0
#> 175   0.582326702      0
#> 176  -0.331119154      0
#> 177   0.904739992      0
#> 178   0.559683975      0
#> 179   0.013587715      0
#> 180   0.388329106      0
#> 181  -0.977254747      0
#> 182  -1.410015373      0
#> 183  -0.131189293      0
#> 184  -0.433368029      0
#> 185   0.925882880      0
#> 186  -1.460643969      0
#> 187   1.110758997      0
#> 188   1.160776441      0
#> 189   1.745842124      0
#> 190   0.979666995      0
#> 191  -0.256483489      0
#> 192   0.894400851      0
#> 193   0.858234144      0
#> 194  -0.290365141      0
#> 195   0.071557535      0
#> 196   0.180009619      0
#> 197   0.368943635      0
#> 198   0.153114182      0
#> 199   0.361924200      0
#> 200  -0.584614752      0
#> 201   0.692522859      0
#> 202  -1.093964972      0
#> 203  -0.413970609      0
#> 204  -0.427723752      0
#> 205   0.559921909      0
#> 206   2.520919633      0
#> 207   0.582206226      0
#> 208   0.660323847      0
#> 209   1.070533447      0
#> 210  -0.070408099      0
#> 211  -1.765712627      0
#> 212  -0.120398788      0
#> 213  -0.030858264      0
#> 214  -1.353414097      0
#> 215   0.760792204      0
#> 216   0.214689771      0
#> 217   0.815771509      0
#> 218   0.131964367      0
#> 219  -0.026601336      0
#> 220  -0.079102272      0
#> 221   0.276897542      0
#> 222  -2.044515124      0
#> 223   0.451703576      0
#> 224   0.145875429      0
#> 225  -1.289725173      0
#> 226   0.703105124      0
#> 227   0.982603242      0
#> 228  -0.996049251      0
#> 229   0.463868930      0
#> 230   0.228546642      0
#> 231  -1.589647393      0
#> 232   0.882543365      0
#> 233  -0.349523976      0
#> 234   0.874081043      0
#> 235   0.427097001      0
#> 236   0.219682329      0
#> 237   2.347960671      0
#> 238   0.328674179      0
#> 239  -0.458190137      0
#> 240   0.782606605      0
#> 241  -0.314186963      0
#> 242   1.783746696      0
#> 243  -0.179565888      0
#> 244   0.784388394      0
#> 245  -0.220059749      0
#> 246   0.829192792      0
#> 247  -1.653748977      0
#> 248   0.482213099      0
#> 249  -0.114533084      0
#> 250   1.671382941      0
#> 251   1.550266484      0
#> 252  -1.225779772      0
#> 253   0.179566727      0
#> 254   0.527356479      0
#> 255  -0.540071728      0
#> 256  -1.140140089      0
#> 257   0.673039596      0
#> 258   1.630477608      0
#> 259   1.820961627      0
#> 260   1.165381928      0
#> 261  -0.796269485      0
#> 262   0.784269334      0
#> 263   0.181347194      0
#> 264  -0.156455608      0
#> 265  -0.326788530      0
#> 266  -0.376044964      0
#> 267   0.679620807      0
#> 268  -0.413842139      0
#> 269  -0.917480916      0
#> 270  -1.345443489      0
#> 271  -1.027958735      0
#> 272  -1.088866562      0
#> 273  -0.489969677      0
#> 274  -0.089752442      0
#> 275   0.641851754      0
#> 276  -0.218109946      0
#> 277   0.157890011      0
#> 278   0.068502001      0
#> 279   0.026855612      0
#> 280   1.447359995      0
#> 281   0.278436160      0
#> 282  -0.117028132      0
#> 283  -0.601290518      0
#> 284  -0.854709931      0
#> 285  -0.525127709      0
#> 286   0.151666841      0
#> 287   2.242268714      0
#> 288   1.190619206      0
#> 289  -0.439113827      0
#> 290  -1.401193252      0
#> 291   0.091329636      0
#> 292  -0.960541921      0
#> 293   0.612886499      0
#> 294  -0.518777812      0
#> 295   0.177098300      0
#> 296   0.140724251      0
#> 297  -1.084127157      0
#> 298   1.287568427      0
#> 299   0.004587373      0
#> 300  -1.201721947      0
#> 301  -2.347862521      0
#> 302   0.650255005      0
#> 303  -1.045389688      0
#> 304  -0.201144167      0
#> 305  -1.144564004      0
#> 306   1.120786694      0
#> 307   0.209977196      0
#> 308   1.154379901      0
#> 309  -2.321733650      0
#> 310   0.529981073      0
#> 311  -0.160046076      0
#> 312  -2.138644343      0
#> 313  -1.001787393      0
#> 314  -2.273334192      0
#> 315   1.072900358      0
#> 316  -0.454300929      0
#> 317   1.418168855      0
#> 318  -1.041933596      0
#> 319  -0.549962049      0
#> 320  -0.495440352      0
#> 321   0.687765177      0
#> 322  -0.307580770      0
#> 323   0.838888616      0
#> 324  -0.090688473      0
#> 325   0.421583445      0
#> 326  -0.548129123      0
#> 327  -0.732395663      0
#> 328  -0.347171589      0
#> 329   0.641448111      0
#> 330   0.541258144      0
#> 331  -0.837536116      0
#> 332  -0.323780275      0
#> 333  -1.139394144      0
#> 334  -1.417218265      0
#> 335  -1.763371147      0
#> 336   1.150840413      0
#> 337   0.628013468      0
#> 338  -0.155323850      0
#> 339   0.315691220      0
#> 340  -0.235318215      0
#> 341  -0.373204550      0
#> 342  -0.171498194      0
#> 343  -0.671459147      0
#> 344  -0.458178844      0
#> 345   0.348654575      0
#> 346   0.657409879      0
#> 347  -0.561407474      0
#> 348  -0.498286566      0
#> 349   0.127311891      0
#> 350  -1.752311577      0
#> 351   0.680502447      0
#> 352  -0.223083480      0
#> 353  -0.512021831      0
#> 354   2.583298521      0
#> 355   1.345949228      0
#> 356  -1.016190141      0
#> 357   0.093217172      0
#> 358  -1.437673528      0
#> 359   1.552037103      0
#> 360   1.559475611      0
#> 361   1.456914784      0
#> 362   0.153759877      0
#> 363   1.692884737      0
#> 364   0.252452794      0
#> 365   1.312553952      0
#> 366  -1.131238819      0
#> 367  -0.939431341      0
#> 368  -1.483757052      0
#> 369   0.914324063      0
#> 370  -1.011934783      0
#> 371   0.265042000      0
#> 372   0.232106948      0
#> 373   1.713172989      0
#> 374  -1.188998476      0
#> 375  -0.552680863      0
#> 376  -1.050782467      0
#> 377  -1.972559156      0
#> 378   0.173345997      0
#> 379   0.600259165      0
#> 380   0.672328893      0
#> 381   0.451776563      0
#> 382  -0.539433272      0
#> 383  -0.973678629      0
#> 384  -0.823751665      0
#> 385   0.678284332      0
#> 386  -1.555711829      0
#> 387  -0.911295144      0
#> 388  -0.514008528      0
#> 389  -0.523946596      0
#> 390   1.033436657      0
#> 391  -1.841559721      0
#> 392   1.425681387      0
#> 393  -0.015992813      0
#> 394   0.367398950      0
#> 395  -0.270843520      0
#> 396  -0.369157568      0
#> 397   1.103758309      0
#> 398   0.060211670      0
#> 399  -0.042486351      0
#> 400  -1.305580969      0
#> 401  -2.609045158      0
#> 402   0.357229258      0
#> 403   1.190386048      0
#> 404   0.097630380      0
#> 405  -0.303023577      0
#> 406   0.278116597      0
#> 407   0.878572084      0
#> 408  -0.635776540      0
#> 409  -0.180942882      0
#> 410  -1.202039792      0
#> 411   0.559545171      0
#> 412  -0.002726708      0
#> 413  -1.369135627      0
#> 414   1.202064242      0
#> 415   1.059390147      0
#> 416   1.035924615      0
#> 417   0.967038883      0
#> 418  -0.966936968      0
#> 419  -2.003107375      0
#> 420  -1.628489634      0
#> 421   0.155159917      0
#> 422  -0.889789139      0
#> 423   1.235393723      0
#> 424  -1.684159871      0
#> 425   0.244267451      0
#> 426   1.268276159      0
#> 427   0.533486878      0
#> 428   1.624771562      0
#> 429   1.627092002      0
#> 430  -0.447854775      0
#> 431  -1.604188862      0
#> 432  -1.065565527      0
#> 433  -0.515247792      0
#> 434  -0.883779743      0
#> 435  -0.574307191      0
#> 436  -0.973875088      0
#> 437   0.921685883      0
#> 438   0.036620224      0
#> 439  -0.110633193      0
#> 440  -1.259348781      0
#> 441   0.285969000      0
#> 442   0.863445155      0
#> 443   0.781482560      0
#> 444   1.363326088      0
#> 445   1.121569346      0
#> 446  -0.015156231      0
#> 447  -0.726081594      0
#> 448  -0.663767469      0
#> 449   0.373641116      0
#> 450  -0.925649318      0
#> 451   0.937605450      0
#> 452  -1.627792516      0
#> 453  -1.392557024      0
#> 454  -0.487797245      0
#> 455   1.350709426      0
#> 456   0.210242579      0
#> 457  -0.209200057      0
#> 458   0.200651299      0
#> 459   0.058951581      0
#> 460  -0.324670031      0
#> 461  -1.653806011      0
#> 462   0.379362453      0
#> 463   1.031525256      0
#> 464   0.636441006      0
#> 465  -0.009707890      0
#> 466  -1.927431274      0
#> 467   2.473442731      0
#> 468   0.274977214      0
#> 469  -0.471654073      0
#> 470   1.656230918      0
#> 471   0.011885031      0
#> 472   0.683225394      0
#> 473  -0.299390032      0
#> 474  -0.323480403      0
#> 475  -1.785692353      0
#> 476  -0.727216943      0
#> 477   0.452186779      0
#> 478   0.801656599      0
#> 479  -0.774776958      0
#> 480  -1.191629588      0
#> 481  -0.078690641      0
#> 482   0.868616862      0
#> 483  -0.337027606      0
#> 484   0.453219277      0
#> 485   0.478312681      0
#> 486  -0.422780159      0
#> 487   1.748779723      0
#> 488  -1.929597056      0
#> 489  -0.851477451      0
#> 490  -0.047399998      0
#> 491  -0.383422025      0
#> 492  -0.613172229      0
#> 493  -0.753216023      0
#> 494  -0.339388550      0
#> 495  -0.386841536      0
#> 496   0.591956753      0
#> 497   0.151286876      0
#> 498  -1.011846414      0
#> 499  -0.513496595      0
#> 500  -2.184101468      0
#> 501   1.888417500      0
#> 502   0.360673822      0
#> 503  -1.658395011      0
#> 504   2.051673184      0
#> 505   0.646656900      0
#> 506  -0.015492570      0
#> 507   1.138534715      0
#> 508  -0.315773973      0
#> 509  -0.352213851      0
#> 510  -2.444287726      0
#> 511  -0.654108502      0
#> 512  -0.001958118      0
#> 513  -0.162456698      0
#> 514   0.661931614      0
#> 515   1.503506489      0
#> 516   0.652689918      0
#> 517   1.497515914      0
#> 518  -3.151010444      0
#> 519   1.057691526      0
#> 520  -1.128289869      0
#> 521   0.884476658      0
#> 522   1.572136102      0
#> 523  -2.607332545      0
#> 524  -3.058914655      0
#> 525  -0.676722388      0
#> 526  -0.627704850      0
#> 527  -1.750113936      0
#> 528   0.899284727      0
#> 529  -0.300299776      0
#> 530  -0.136373114      0
#> 531   1.353782871      0
#> 532  -0.463412503      0
#> 533  -0.602434678      0
#> 534  -0.011640219      0
#> 535   0.593632585      0
#> 536  -1.214744898      0
#> 537   0.170944838      0
#> 538  -0.909578975      0
#> 539  -0.763096303      0
#> 540   2.034673170      0
#> 541  -1.343129202      0
#> 542  -0.327627147      0
#> 543   1.054212925      0
#> 544  -0.050469993      0
#> 545  -0.192430145      0
#> 546  -0.968373071      0
#> 547   0.586131133      0
#> 548  -1.657924491      0
#> 549   0.784874191      0
#> 550   0.346115853      0
#> 551  -1.412817147      0
#> 552  -0.397883784      0
#> 553  -0.992737731      0
#> 554   0.846876084      0
#> 555  -0.974197964      0
#> 556  -0.024812490      0
#> 557   0.185667022      0
#> 558  -0.145413211      0
#> 559  -0.500326431      0
#> 560  -0.002859519      0
#> 561   0.672262446      0
#> 562   0.951124042      0
#> 563   0.087951848      0
#> 564   1.093959870      0
#> 565  -0.051905887      0
#> 566  -1.675888954      0
#> 567  -1.286953127      0
#> 568   0.403279688      0
#> 569  -1.579180690      0
#> 570  -0.777306661      0
#> 571   0.117283672      0
#> 572  -0.862473634      0
#> 573   0.103977180      0
#> 574  -1.868502021      0
#> 575   1.189996416      0
#> 576  -0.207058324      0
#> 577  -2.838558245      0
#> 578   1.470220042      0
#> 579  -1.085979439      0
#> 580  -0.243648678      0
#> 581  -1.216604149      0
#> 582  -1.471274935      0
#> 583  -0.649628388      0
#> 584  -0.604719966      0
#> 585   1.244160621      0
#> 586   0.761537640      0
#> 587  -1.207501378      0
#> 588   0.556843290      0
#> 589  -1.498008738      0
#> 590   0.399604264      0
#> 591   0.666251159      0
#> 592   0.034212602      0
#> 593  -1.748427044      0
#> 594  -0.404683974      0
#> 595  -0.352477921      0
#> 596  -1.417499532      0
#> 597  -0.123527257      0
#> 598  -0.839199107      0
#> 599   1.445409773      0
#> 600  -1.731923313      0
#> 601   0.029113445      0
#> 602   0.521245826      0
#> 603  -0.123852098      0
#> 604  -0.215069280      0
#> 605   0.545444770      0
#> 606   0.212905303      0
#> 607  -0.717600968      0
#> 608  -2.082353623      0
#> 609   0.488484457      0
#> 610   0.845050419      0
#> 611  -0.514497017      0
#> 612   0.371405070      0
#> 613  -0.867508163      0
#> 614   1.187733535      0
#> 615  -1.030126862      0
#> 616  -1.258490654      0
#> 617   0.608733540      0
#> 618   0.197438750      0
#> 619   0.033648476      0
#> 620   0.085411428      0
#> 621  -0.226189364      0
#> 622   1.489008778      0
#> 623  -0.846401760      0
#> 624   0.931321611      0
#> 625   0.008403198      0
#> 626  -0.819749862      0
#> 627   0.726321999      0
#> 628   0.858057108      0
#> 629   0.730757140      0
#> 630   1.872458066      0
#> 631   0.113910119      0
#> 632   1.953661767      0
#> 633   1.140337359      0
#> 634   1.971502108      0
#> 635   0.412685786      0
#> 636   0.710773317      0
#> 637   0.662543301      0
#> 638   0.629895717      0
#> 639  -0.763250292      0
#> 640   1.364337448      0
#> 641   0.033746389      0
#> 642  -3.367111421      0
#> 643   0.466310378      0
#> 644   0.306559772      0
#> 645  -1.358921982      0
#> 646  -0.436277883      0
#> 647   1.038257221      0
#> 648  -0.049688075      0
#> 649   0.783621655      0
#> 650  -0.723642854      0
#> 651   0.445627091      0
#> 652  -0.115585942      0
#> 653   0.467082299      0
#> 654  -0.084762053      0
#> 655   0.694955372      0
#> 656  -0.778773083      0
#> 657  -1.056135360      0
#> 658  -0.174638311      0
#> 659   0.342509144      0
#> 660  -0.060328654      0
#> 661   0.492954608      0
#> 662   1.186335847      0
#> 663  -0.793527554      0
#> 664  -0.851251929      0
#> 665   0.603273361      0
#> 666   0.641967967      0
#> 667  -0.054947411      0
#> 668   0.356975223      0
#> 669   1.008252305      0
#> 670   0.593534525      0
#> 671   0.618428458      0
#> 672   0.183563299      0
#> 673  -1.498631423      0
#> 674  -0.996916706      0
#> 675   0.969935342      0
#> 676   0.511253879      0
#> 677   1.768367820      0
#> 678  -0.493269104      0
#> 679   2.196663467      0
#> 680  -1.557328577      0
#> 681   0.142809444      0
#> 682   0.910053657      0
#> 683   0.916438911      0
#> 684   0.797340995      0
#> 685   0.079836748      0
#> 686  -0.492041887      0
#> 687   1.913900427      0
#> 688  -0.235936813      0
#> 689  -0.577548955      0
#> 690   0.828195027      0
#> 691   0.013992826      0
#> 692   0.141795289      0
#> 693   0.835767402      0
#> 694   0.316947070      0
#> 695   1.026773327      0
#> 696  -0.652954107      0
#> 697  -0.394915224      0
#> 698  -0.739107041      0
#> 699   0.668614451      0
#> 700   0.374997488      0
#> 701  -2.409965720      0
#> 702  -1.527904098      0
#> 703   0.292895131      0
#> 704  -0.284094197      0
#> 705   1.645875791      0
#> 706   0.102831221      0
#> 707  -0.869273650      0
#> 708  -2.453714457      0
#> 709  -0.511986253      0
#> 710  -1.291752359      0
#> 711  -0.956260209      0
#> 712   1.213420570      0
#> 713  -1.502174487      0
#> 714   0.693136105      0
#> 715   1.060050210      0
#> 716  -0.167889930      0
#> 717   0.448100082      0
#> 718  -0.326731261      0
#> 719   0.527959048      0
#> 720   0.382362864      0
#> 721   0.458089424      0
#> 722   0.352388822      0
#> 723   1.453167360      0
#> 724   1.003213538      0
#> 725  -0.668052319      0
#> 726  -0.662142231      0
#> 727   0.545824162      0
#> 728  -1.109138552      0
#> 729   1.971092672      0
#> 730   0.261796402      0
#> 731  -3.257601482      0
#> 732  -0.203846826      0
#> 733  -1.252629900      0
#> 734   0.533992031      0
#> 735  -0.130701122      0
#> 736   1.357081826      0
#> 737  -1.756055389      0
#> 738  -0.692936963      0
#> 739   1.127767145      0
#> 740   0.497103943      0
#> 741   0.059821569      0
#> 742  -0.107883241      0
#> 743   0.285216428      0
#> 744   0.379898203      0
#> 745  -0.859455098      0
#> 746   0.616817690      0
#> 747   2.016154995      0
#> 748   0.175379529      0
#> 749   2.098174891      0
#> 750   2.645001369      0
#> 751  -1.106546535      0
#> 752   0.005841722      0
#> 753   0.489063071      0
#> 754   0.339850029      0
#> 755   0.449796496      0
#> 756   0.274073776      0
#> 757   0.192925054      0
#> 758   0.031516204      0
#> 759   0.489110860      0
#> 760   1.310096418      0
#> 761  -0.298739400      0
#> 762  -0.311462188      0
#> 763   0.144846240      0
#> 764   2.143450519      0
#> 765   0.241174782      0
#> 766   0.287787566      0
#> 767  -3.383283985      0
#> 768  -2.617360843      0
#> 769  -0.878045956      0
#> 770   0.267275957      0
#> 771  -0.277497702      0
#> 772   0.698647714      0
#> 773  -1.244232922      0
#> 774  -1.738592030      0
#> 775  -0.708949535      0
#> 776   1.737069345      0
#> 777  -0.360604574      0
#> 778   1.087714848      0
#> 779  -2.428762883      0
#> 780  -1.844303154      0
#> 781   0.015559855      0
#> 782  -0.142768823      0
#> 783   1.987852265      0
#> 784  -0.569560808      0
#> 785  -1.330571904      0
#> 786   0.334933725      0
#> 787  -0.068891904      0
#> 788   0.617233914      0
#> 789   0.857345508      0
#> 790   0.067792151      0
#> 791  -0.182991754      0
#> 792  -0.537994807      0
#> 793  -1.301060472      0
#> 794  -1.008952694      0
#> 795  -0.329047816      0
#> 796   0.233452949      0
#> 797   1.134238751      0
#> 798   1.269616482      0
#> 799  -0.935232067      0
#> 800  -1.662752357      0
#> 801  -0.544591667      0
#> 802  -0.851302735      0
#> 803  -0.142284300      0
#> 804  -0.622637078      0
#> 805   0.951729070      0
#> 806  -0.201800302      0
#> 807  -0.720650632      0
#> 808   0.574253047      0
#> 809   0.632445069      0
#> 810  -0.972639441      0
#> 811   1.123026700      0
#> 812   0.256522558      0
#> 813   0.005918962      0
#> 814   0.199356000      0
#> 815  -1.674062131      0
#> 816  -1.229865501      0
#> 817  -0.231643143      0
#> 818   0.103902049      0
#> 819  -0.248925849      0
#> 820   1.349798649      0
#> 821  -0.071205414      0
#> 822  -0.421099281      0
#> 823   0.423107571      0
#> 824  -0.096496101      0
#> 825  -0.986731796      0
#> 826  -0.989253116      0
#> 827  -0.253774825      0
#> 828   1.490724908      0
#> 829  -0.919738479      0
#> 830   0.214245472      0
#> 831  -0.226605742      0
#> 832  -0.336490607      0
#> 833  -2.217768324      0
#> 834  -0.217039336      0
#> 835   0.610486384      0
#> 836  -0.258953167      0
#> 837   0.156374723      0
#> 838   0.888566815      0
#> 839   0.517284190      0
#> 840   0.086260968      0
#> 841   0.982089662      0
#> 842  -1.864867754      0
#> 843  -0.811631155      0
#> 844  -0.073215138      0
#> 845   0.898268828      0
#> 846  -0.555746818      0
#> 847  -0.766981370      0
#> 848   0.341974874      0
#> 849  -0.421459877      0
#> 850  -1.361413608      0
#> 851  -1.251681768      0
#> 852  -1.222440692      0
#> 853  -0.621601290      0
#> 854  -0.046613035      0
#> 855   0.816491917      0
#> 856   0.625101428      0
#> 857   0.938426923      0
#> 858   0.605541019      0
#> 859   1.635500687      0
#> 860   1.204143197      0
#> 861  -0.836800628      0
#> 862   0.424600985      0
#> 863  -2.347000819      0
#> 864   1.740541904      0
#> 865   1.069651643      0
#> 866   0.416639392      0
#> 867  -0.725653441      0
#> 868  -0.112507009      0
#> 869  -0.199223311      0
#> 870   0.076280396      0
#> 871   0.124530579      0
#> 872  -0.635784300      0
#> 873   0.388340945      0
#> 874  -0.396765656      0
#> 875  -0.129313532      0
#> 876   0.880236656      0
#> 877  -0.803836343      0
#> 878  -2.270385992      0
#> 879   0.430088994      0
#> 880  -1.109998900      0
#> 881   0.137108538      0
#> 882   0.343990313      0
#> 883  -1.135498936      0
#> 884  -2.331790650      0
#> 885   0.530128683      0
#> 886   0.024907957      0
#> 887  -0.522979359      0
#> 888   0.096283587      0
#> 889  -0.890465686      0
#> 890  -1.115580591      0
#> 891  -0.444519168      0
#> 892  -0.137769958      0
#> 893   1.483786700      0
#> 894  -1.157811973      0
#> 895   2.173360491      0
#> 896  -1.623876832      0
#> 897   1.710164469      0
#> 898   1.030909172      0
#> 899   0.513548162      0
#> 900   0.255948369      0
#> 901   0.172996924      0
#> 902  -1.194095854      0
#> 903  -1.127596578      0
#> 904  -1.733882889      0
#> 905   0.313586615      0
#> 906  -1.727487122      0
#> 907  -0.171092284      0
#> 908  -0.383350708      0
#> 909   0.920460787      0
#> 910  -0.137506572      0
#> 911   0.261506348      0
#> 912  -0.267902735      0
#> 913   0.186367891      0
#> 914   1.372895098      0
#> 915   0.154647092      0
#> 916   1.162659106      1
#> 917   0.376827682      1
#> 918   0.337727067      1
#> 919   0.382452425      1
#> 920   2.334149985      1
#> 921   0.186895042      1
#> 922  -0.384233622      1
#> 923   1.703622557      1
#> 924   1.082059569      1
#> 925  -0.128570541      1
#> 926   1.331078163      1
#> 927   0.949284983      1
#> 928  -0.176513991      1
#> 929   0.610402563      1
#> 930  -0.404535559      1
#> 931   1.059198770      1
#> 932   0.052826148      1
#> 933  -0.079892032      1
#> 934   0.867022048      1
#> 935   1.953975783      1
#> 936   2.497040103      1
#> 937   1.308124351      1
#> 938   0.273629853      1
#> 939   1.412771995      1
#> 940  -0.804506673      1
#> 941  -0.478449045      1
#> 942   1.928531496      1
#> 943   4.047611699      1
#> 944   0.287759678      1
#> 945   0.705600251      1
#> 946   1.365546410      1
#> 947   0.683132405      1
#> 948   1.832873180      1
#> 949   0.646865491      1
#> 950   0.431458616      1
#> 951   2.462838343      1
#> 952   1.466427240      1
#> 953   1.544771647      1
#> 954   0.491666087      1
#> 955   1.534505825      1
#> 956   2.181801589      1
#> 957  -0.055423517      1
#> 958   0.906214412      1
#> 959   1.262983601      1
#> 960   2.063983276      1
#> 961   0.151581477      1
#> 962   2.603046663      1
#> 963   1.840006041      1
#> 964   0.039493054      1
#> 965  -0.784675608      1
#> 966  -0.638305295      1
#> 967   0.774571925      1
#> 968   0.599938021      1
#> 969  -0.629317394      1
#> 970   0.116779874      1
#> 971   0.390800841      1
#> 972  -0.257134144      1
#> 973   2.231801299      1
#> 974   1.788491128      1
#> 975  -0.513460301      1
#> 976   0.150187644      1
#> 977   1.324915024      1
#> 978   0.481410070      1
#> 979   1.593616057      1
#> 980   2.709409864      1
#> 981  -0.401802633      1
#> 982   1.475298612      1
#> 983   1.459637713      1
#> 984  -0.827069986      1
#> 985  -0.974186106      1
#> 986  -0.232176277      1
#> 987  -0.387549809      1
#> 988  -1.475160375      1
#> 989   0.791374339      1
#> 990  -1.515814059      1
#> 991   2.038601403      1
#> 992  -0.007360827      1
#> 993   1.423037231      1
#> 994   0.486032958      1
#> 995   0.489175918      1
#> 996   0.466872660      1
#> 997  -0.805537193      1
#> 998   2.427299454      1
#> 999   0.068269074      1
#> 1000  0.877215050      1