Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     2.529244005      0
#> 2    -0.687622119      0
#> 3     0.683975852      0
#> 4     0.115406430      0
#> 5     1.100477486      0
#> 6    -1.066754336      0
#> 7    -1.189722049      0
#> 8    -0.992954316      0
#> 9     0.611988667      0
#> 10    2.535423835      0
#> 11    0.475616878      0
#> 12   -0.260513902      0
#> 13    2.092143532      0
#> 14    0.190808998      0
#> 15    0.949098130      0
#> 16   -0.154286096      0
#> 17    0.817234912      0
#> 18   -0.294037300      0
#> 19    0.338404761      0
#> 20   -0.447851774      0
#> 21    0.816590571      0
#> 22    0.602433278      0
#> 23    1.245278094      0
#> 24    0.343871581      0
#> 25    0.924004691      0
#> 26    1.621926655      0
#> 27   -1.242607363      0
#> 28    0.771277907      0
#> 29    1.174210365      0
#> 30    0.097310085      0
#> 31   -0.987282119      0
#> 32   -1.049206859      0
#> 33   -1.188135338      0
#> 34    0.768064757      0
#> 35   -0.562837429      0
#> 36   -1.215015741      0
#> 37    2.232065687      0
#> 38   -0.649335397      0
#> 39    0.103422681      0
#> 40   -0.232499937      0
#> 41    0.230377302      0
#> 42    0.508034165      0
#> 43    0.392328365      0
#> 44    2.011863605      0
#> 45    0.608017545      0
#> 46    0.469667626      0
#> 47   -0.183183713      0
#> 48    0.528156367      0
#> 49   -1.072557020      0
#> 50   -0.545949839      0
#> 51    0.436951960      0
#> 52   -0.258795034      0
#> 53   -0.276891081      0
#> 54    0.601385967      0
#> 55   -1.051019270      0
#> 56    0.532114682      0
#> 57    0.508657082      0
#> 58    0.686251455      0
#> 59   -2.315483696      0
#> 60    0.831962450      0
#> 61    1.242329977      0
#> 62    1.676664327      0
#> 63   -0.158230263      0
#> 64   -0.464291180      0
#> 65    0.578258984      0
#> 66    0.732535626      0
#> 67    0.392831102      0
#> 68   -0.213000407      0
#> 69   -0.926811177      0
#> 70    0.001396988      0
#> 71   -1.000517556      0
#> 72   -1.303706493      0
#> 73    0.528112259      0
#> 74    0.880289041      0
#> 75    0.779938519      0
#> 76    0.737758786      0
#> 77    1.775748425      0
#> 78   -0.820101297      0
#> 79    0.981307549      0
#> 80    0.566977852      0
#> 81    0.341132958      0
#> 82   -0.237202179      0
#> 83    0.102241620      0
#> 84    1.022824486      0
#> 85   -1.366662476      0
#> 86   -1.536760340      0
#> 87    0.170072028      0
#> 88   -1.442794569      0
#> 89   -1.041752848      0
#> 90    0.243363306      0
#> 91    2.028291599      0
#> 92   -0.422602166      0
#> 93   -0.426997647      0
#> 94    0.235622398      0
#> 95    1.048511733      0
#> 96    0.531642450      0
#> 97   -0.112335512      0
#> 98    0.556719625      0
#> 99   -0.121589557      0
#> 100  -1.091743860      0
#> 101   0.995560660      0
#> 102  -1.506727095      0
#> 103  -0.826855487      0
#> 104   0.074840053      0
#> 105   1.009576510      0
#> 106   0.714721319      0
#> 107  -0.874167972      0
#> 108   0.521811194      0
#> 109   0.589020987      0
#> 110  -0.329495850      0
#> 111   0.527789991      0
#> 112  -0.809923570      0
#> 113   0.724884306      0
#> 114   0.258405913      0
#> 115   0.014672193      0
#> 116   0.885385699      0
#> 117  -0.085618558      0
#> 118  -1.595602421      0
#> 119  -1.490908746      0
#> 120   2.348858945      0
#> 121   1.705877216      0
#> 122   0.026855146      0
#> 123  -3.598851386      0
#> 124  -0.380986250      0
#> 125  -0.233050204      0
#> 126  -0.573468106      0
#> 127   0.068975303      0
#> 128   0.840716201      0
#> 129  -0.414869641      0
#> 130   1.177291989      0
#> 131  -0.111944427      0
#> 132   1.310122750      0
#> 133  -1.235296723      0
#> 134  -0.165368148      0
#> 135  -0.481120853      0
#> 136  -0.360614236      0
#> 137  -0.183192083      0
#> 138  -0.325868046      0
#> 139   1.443138614      0
#> 140  -0.396128626      0
#> 141  -0.014867915      0
#> 142   1.579266808      0
#> 143   0.525831820      0
#> 144  -0.714806142      0
#> 145  -0.209402998      0
#> 146  -0.333734417      0
#> 147  -1.624334415      0
#> 148  -0.668177129      0
#> 149   1.222704833      0
#> 150   1.905899050      0
#> 151   1.777831602      0
#> 152  -1.186792541      0
#> 153   0.829252434      0
#> 154   0.682828489      0
#> 155   0.600716919      0
#> 156  -1.229163631      0
#> 157   0.231054684      0
#> 158   0.316426788      0
#> 159  -0.900555684      0
#> 160   1.597768623      0
#> 161   0.854565316      0
#> 162   1.827670410      0
#> 163   1.245589359      0
#> 164  -0.643440759      0
#> 165  -1.876749279      0
#> 166  -0.391429705      0
#> 167  -1.409496515      0
#> 168  -1.076637564      0
#> 169   0.997631985      0
#> 170   0.168297084      0
#> 171   0.286750787      0
#> 172   2.078412700      0
#> 173   0.191062239      0
#> 174  -0.412074306      0
#> 175  -0.648992090      0
#> 176  -0.051392774      0
#> 177   0.565538887      0
#> 178   0.053198994      0
#> 179  -0.242687826      0
#> 180   1.270378045      0
#> 181  -0.322239920      0
#> 182   0.071628004      0
#> 183  -0.723188909      0
#> 184   0.205612763      0
#> 185  -0.253929530      0
#> 186   0.203363754      0
#> 187   0.132174504      0
#> 188   1.176485879      0
#> 189  -1.122625276      0
#> 190  -1.026537177      0
#> 191  -0.839316805      0
#> 192   0.604906208      0
#> 193   0.165376466      0
#> 194   0.918202925      0
#> 195  -0.791872445      0
#> 196   0.372429500      0
#> 197   1.139580790      0
#> 198   0.119979323      0
#> 199  -0.916586041      0
#> 200   0.321511184      0
#> 201   0.678763870      0
#> 202   0.124654116      0
#> 203   0.463474874      0
#> 204   1.392049921      0
#> 205   0.800647098      0
#> 206   0.003590674      0
#> 207   0.264434215      0
#> 208   0.185181770      0
#> 209  -0.119202024      0
#> 210  -0.205513531      0
#> 211  -0.621262109      0
#> 212  -0.799144216      0
#> 213  -1.568419657      0
#> 214  -0.818451132      0
#> 215  -2.035254944      0
#> 216   2.700038502      0
#> 217  -0.727208638      0
#> 218  -1.378105858      0
#> 219  -0.157176860      0
#> 220  -0.191864726      0
#> 221   0.001338142      0
#> 222   1.502514431      0
#> 223   0.588877936      0
#> 224   0.521599030      0
#> 225   0.951478110      0
#> 226  -0.086838024      0
#> 227  -0.730615366      0
#> 228   0.269565183      0
#> 229   0.081563579      0
#> 230   0.232190326      0
#> 231   1.474505262      0
#> 232  -0.398613410      0
#> 233  -0.131918376      0
#> 234   0.214230498      0
#> 235   0.025776410      0
#> 236   1.253125849      0
#> 237  -1.281537080      0
#> 238  -0.856420281      0
#> 239  -0.351802762      0
#> 240   1.189658406      0
#> 241  -0.645791325      0
#> 242  -0.245035640      0
#> 243   1.713060156      0
#> 244   0.929324090      0
#> 245  -0.055089571      0
#> 246   0.407158501      0
#> 247   1.263513087      0
#> 248  -0.454186081      0
#> 249  -1.949712835      0
#> 250  -1.464305652      0
#> 251  -1.089088558      0
#> 252   0.485237030      0
#> 253  -1.216635541      0
#> 254  -0.493067003      0
#> 255  -1.521229418      0
#> 256   2.008862140      0
#> 257  -0.044797695      0
#> 258   1.127670601      0
#> 259   0.446508014      0
#> 260   0.393779636      0
#> 261  -0.001464962      0
#> 262   1.925678303      0
#> 263   0.770620307      0
#> 264   0.283008828      0
#> 265   0.306559505      0
#> 266   1.536602291      0
#> 267  -1.181723787      0
#> 268   0.877156956      0
#> 269  -0.488831890      0
#> 270  -0.089349830      0
#> 271   1.079134669      0
#> 272  -0.206364685      0
#> 273   1.709206579      0
#> 274  -0.463480711      0
#> 275   1.179763919      0
#> 276   0.667125926      0
#> 277  -0.866596709      0
#> 278  -1.587765095      0
#> 279   1.491656228      0
#> 280  -0.713659429      0
#> 281   0.693138022      0
#> 282   1.199049830      0
#> 283   0.154401971      0
#> 284  -0.526889385      0
#> 285   0.967118609      0
#> 286   0.675406910      0
#> 287  -1.423318676      0
#> 288   0.629543588      0
#> 289   1.417914508      0
#> 290  -0.963501642      0
#> 291  -1.349860842      0
#> 292  -0.740971284      0
#> 293  -0.365055262      0
#> 294   0.163584689      0
#> 295   1.148617862      0
#> 296   1.038382226      0
#> 297   1.088876812      0
#> 298  -0.841705934      0
#> 299   0.106742885      0
#> 300   0.180345969      0
#> 301   0.042157676      0
#> 302  -1.222034979      0
#> 303   0.400502519      0
#> 304   2.033852599      0
#> 305  -0.283990943      0
#> 306   0.265930427      0
#> 307  -1.108850742      0
#> 308  -0.252566373      0
#> 309   0.040242281      0
#> 310  -0.462760306      0
#> 311   0.659087451      0
#> 312  -0.454512910      0
#> 313  -0.673852748      0
#> 314  -1.005592752      0
#> 315   0.363343595      0
#> 316  -0.277227588      0
#> 317  -0.640718784      0
#> 318   0.253013433      0
#> 319  -1.387261787      0
#> 320  -0.874573653      0
#> 321  -0.219323645      0
#> 322   2.074923602      0
#> 323   0.697005413      0
#> 324  -2.480639857      0
#> 325   0.711092362      0
#> 326  -1.618395422      0
#> 327   0.688862486      0
#> 328   0.671160132      0
#> 329   0.903560858      0
#> 330   0.110141508      0
#> 331  -0.396983394      0
#> 332  -0.867559375      0
#> 333  -0.445857162      0
#> 334   2.656290989      0
#> 335   0.633290683      0
#> 336  -0.103126126      0
#> 337  -0.329720380      0
#> 338   0.812373692      0
#> 339   0.378484789      0
#> 340   0.177265232      0
#> 341   1.524952507      0
#> 342   1.121284605      0
#> 343  -0.355609233      0
#> 344   0.038403983      0
#> 345   0.734683316      0
#> 346   1.107661545      0
#> 347  -1.383962449      0
#> 348   0.842323142      0
#> 349   0.136741016      0
#> 350  -0.674403885      0
#> 351  -0.534239599      0
#> 352  -1.379747773      0
#> 353   2.295182419      0
#> 354   0.121694532      0
#> 355   0.908281067      0
#> 356  -0.011489170      0
#> 357   0.412134899      0
#> 358  -1.188356278      0
#> 359   0.602003482      0
#> 360  -0.734387486      0
#> 361   1.052230248      0
#> 362   0.498467452      0
#> 363   0.408176454      0
#> 364  -1.470696707      0
#> 365  -1.306916178      0
#> 366   0.243730638      0
#> 367  -1.277635155      0
#> 368   1.249420697      0
#> 369   1.878564688      0
#> 370  -0.369043616      0
#> 371   0.328357815      0
#> 372   2.500598889      0
#> 373   1.038831864      0
#> 374  -0.924159612      0
#> 375  -1.777575588      0
#> 376   1.088232092      0
#> 377  -0.526646244      0
#> 378   2.340040316      0
#> 379   0.872958419      0
#> 380   0.300592274      0
#> 381   0.728640221      0
#> 382  -0.123499980      0
#> 383  -0.068838819      0
#> 384   0.135887912      0
#> 385   0.292102787      0
#> 386  -0.690220419      0
#> 387  -1.970027956      0
#> 388   0.415076420      0
#> 389  -1.569038466      0
#> 390  -0.028336218      0
#> 391  -1.197752619      0
#> 392   0.845844608      0
#> 393   0.516423677      0
#> 394   0.009898878      0
#> 395  -1.183737859      0
#> 396  -0.056525397      0
#> 397  -0.042289320      0
#> 398   1.481424768      0
#> 399   0.164205372      0
#> 400   1.690755658      0
#> 401  -0.651687326      0
#> 402  -0.095731817      0
#> 403  -0.458878028      0
#> 404  -0.346386549      0
#> 405  -0.124237384      0
#> 406  -0.933537261      0
#> 407   0.318657938      0
#> 408   0.183749996      0
#> 409  -1.630044348      0
#> 410   2.033089868      0
#> 411  -0.196736384      0
#> 412  -0.804086035      0
#> 413   1.055845587      0
#> 414  -0.160728174      0
#> 415  -0.882041675      0
#> 416   1.027093013      0
#> 417  -0.186629985      0
#> 418   0.209665322      0
#> 419  -1.022046168      0
#> 420   0.532824558      0
#> 421   0.021965456      0
#> 422   0.983451414      0
#> 423  -0.429634459      0
#> 424  -0.682057880      0
#> 425  -1.096667472      0
#> 426   1.194528724      0
#> 427   0.025952435      0
#> 428  -2.445420877      0
#> 429  -0.646910681      0
#> 430  -1.361736433      0
#> 431  -0.507874942      0
#> 432   0.090424515      0
#> 433   0.003766615      0
#> 434   0.038043829      0
#> 435   0.450906557      0
#> 436   0.489389164      0
#> 437  -0.056960976      0
#> 438   1.345683054      0
#> 439  -1.026630554      0
#> 440  -0.151145324      0
#> 441  -0.636169599      0
#> 442  -0.633925554      0
#> 443  -0.738378146      0
#> 444   0.715676265      0
#> 445   0.601191360      0
#> 446  -0.322154031      0
#> 447   0.443164155      0
#> 448   1.401927327      0
#> 449   0.600078138      0
#> 450  -0.688400905      0
#> 451  -1.392452427      0
#> 452  -0.687703006      0
#> 453  -0.608019149      0
#> 454  -0.785166744      0
#> 455  -0.174768287      0
#> 456   0.327739638      0
#> 457   1.825038623      0
#> 458   0.485979174      0
#> 459   1.579905726      0
#> 460  -1.067439046      0
#> 461   0.365941832      0
#> 462   0.111901077      0
#> 463   0.843601708      0
#> 464   0.677153409      0
#> 465   1.894950911      0
#> 466   1.302439129      0
#> 467  -0.335201226      0
#> 468  -0.633365545      0
#> 469  -0.576017060      0
#> 470   0.599826269      0
#> 471   0.619539729      0
#> 472  -0.189986213      0
#> 473   1.576758040      0
#> 474   0.321561159      0
#> 475   0.166127499      0
#> 476  -1.417396643      0
#> 477  -1.289202397      0
#> 478   0.246575993      0
#> 479   0.513247085      0
#> 480   1.386259052      0
#> 481   0.911960343      0
#> 482   0.420456515      0
#> 483   0.907558170      0
#> 484   1.185822268      0
#> 485   1.060931167      0
#> 486   0.656675258      0
#> 487   0.834081695      0
#> 488   0.321074274      0
#> 489   0.469044308      0
#> 490   0.716614993      0
#> 491   0.685464666      0
#> 492   0.740600052      0
#> 493   0.252216601      0
#> 494  -0.345846551      0
#> 495  -0.092514819      0
#> 496   1.041973308      0
#> 497  -2.081195396      0
#> 498   1.157559235      0
#> 499   0.795364390      0
#> 500  -0.559775713      0
#> 501   0.214879014      0
#> 502   0.948621589      0
#> 503  -1.584995960      0
#> 504   0.056405298      0
#> 505  -1.706323451      0
#> 506   0.327776132      0
#> 507   0.224205886      0
#> 508   2.014711803      0
#> 509  -0.476676748      0
#> 510   0.584617506      0
#> 511  -0.461797552      0
#> 512  -0.100982859      0
#> 513  -0.070238119      0
#> 514   0.652348969      0
#> 515   0.590443133      0
#> 516  -0.112380521      0
#> 517   0.672554072      0
#> 518  -0.139637983      0
#> 519   0.596091076      0
#> 520  -0.659325408      0
#> 521  -0.586485453      0
#> 522  -0.020496585      0
#> 523   2.569030628      0
#> 524  -2.144709299      0
#> 525  -0.416005293      0
#> 526   0.600948697      0
#> 527   0.687992220      0
#> 528  -1.274451464      0
#> 529  -0.944507515      0
#> 530   0.697700681      0
#> 531  -1.237812737      0
#> 532  -0.155582549      0
#> 533  -0.971686276      0
#> 534  -0.566230565      0
#> 535  -1.023192245      0
#> 536  -1.784135507      0
#> 537  -0.723081667      0
#> 538   0.240185079      0
#> 539  -0.064315446      0
#> 540   0.828253707      0
#> 541  -1.625858659      0
#> 542   0.706190007      0
#> 543   0.335269300      0
#> 544  -0.510076829      0
#> 545   2.097162524      0
#> 546   0.296270150      0
#> 547   0.823592769      0
#> 548  -1.299961733      0
#> 549   0.062756448      0
#> 550  -1.432202434      0
#> 551   0.656459869      0
#> 552  -1.779114222      0
#> 553   0.090918325      0
#> 554   1.640377933      0
#> 555   0.764356779      0
#> 556  -0.176333196      0
#> 557  -2.414693589      0
#> 558  -0.076400114      0
#> 559  -0.141597211      0
#> 560  -0.658722353      0
#> 561   0.330881706      0
#> 562   0.550596686      0
#> 563   0.204906739      0
#> 564   0.379748813      0
#> 565  -0.837081411      0
#> 566  -0.762337006      0
#> 567   0.394933865      0
#> 568  -0.435474176      0
#> 569  -1.318465982      0
#> 570  -0.032156822      0
#> 571  -0.033642168      0
#> 572  -0.401591906      0
#> 573   1.990141569      0
#> 574   1.314367753      0
#> 575   0.436442919      0
#> 576   0.551429925      0
#> 577   0.665976373      0
#> 578  -1.573256716      0
#> 579  -1.304335258      0
#> 580  -0.826740098      0
#> 581  -1.069087351      0
#> 582   1.303757284      0
#> 583  -0.589393547      0
#> 584   0.029706454      0
#> 585   0.417980681      0
#> 586  -0.345298284      0
#> 587  -0.115738090      0
#> 588  -0.062991226      0
#> 589   0.004491082      0
#> 590  -1.066175693      0
#> 591  -0.575022432      0
#> 592   0.231744888      0
#> 593  -0.266457718      0
#> 594  -0.082753241      0
#> 595  -0.520261748      0
#> 596  -1.139694336      0
#> 597   1.025982536      0
#> 598  -1.317152447      0
#> 599  -0.740836181      0
#> 600   1.364908307      0
#> 601   0.800188425      0
#> 602  -1.291747265      0
#> 603  -0.020361093      0
#> 604  -0.613597411      0
#> 605  -1.034059325      0
#> 606   0.833527839      0
#> 607   1.330398049      0
#> 608   0.934711822      0
#> 609  -0.052995705      0
#> 610  -0.781040661      0
#> 611  -1.138091647      0
#> 612  -0.219104274      0
#> 613  -0.093886039      0
#> 614  -0.022808355      0
#> 615   0.175843812      0
#> 616   1.187290241      0
#> 617   1.304238129      0
#> 618  -1.122028874      0
#> 619  -0.700180674      0
#> 620  -0.661492853      0
#> 621   1.034963104      0
#> 622  -0.514371139      0
#> 623   2.269767422      0
#> 624   0.254380730      0
#> 625  -0.730941745      0
#> 626  -0.985333562      0
#> 627   0.130204730      0
#> 628  -1.948781095      0
#> 629   1.275735503      0
#> 630   0.819182362      0
#> 631  -0.961501017      0
#> 632   0.903973827      0
#> 633  -1.376102883      0
#> 634  -0.456414146      0
#> 635  -0.829164749      0
#> 636  -0.667467258      0
#> 637  -0.574882684      0
#> 638   0.443042307      0
#> 639  -1.299304015      0
#> 640  -1.227526051      0
#> 641   0.012425568      0
#> 642  -1.176869548      0
#> 643  -1.077897969      0
#> 644  -0.338784706      0
#> 645   0.221951495      0
#> 646   0.543216339      0
#> 647  -0.816694167      0
#> 648   0.506406650      0
#> 649   1.693255164      0
#> 650  -0.229937746      0
#> 651   0.350815961      0
#> 652   0.938704831      0
#> 653  -1.174906849      0
#> 654  -2.441361026      0
#> 655   0.599061898      0
#> 656  -1.949473219      0
#> 657   1.514173682      0
#> 658   0.498420947      0
#> 659  -0.481337077      0
#> 660  -0.331139232      0
#> 661   1.121398060      0
#> 662  -0.053675631      0
#> 663   0.158125084      0
#> 664  -0.923503866      0
#> 665  -1.563620398      0
#> 666   0.155987664      0
#> 667  -0.615390037      0
#> 668  -2.319690820      0
#> 669   0.428672610      0
#> 670  -0.775241973      0
#> 671  -1.716269294      0
#> 672  -1.278366049      0
#> 673   0.787035990      0
#> 674  -0.542982049      0
#> 675  -0.609070763      0
#> 676   0.417833475      0
#> 677   0.835840678      0
#> 678   1.325892350      0
#> 679  -2.345328724      0
#> 680  -0.531329155      0
#> 681   0.058750773      0
#> 682   0.731559722      0
#> 683  -0.050649678      0
#> 684  -0.859332549      0
#> 685  -0.446465857      0
#> 686  -3.029530124      0
#> 687   1.071206178      0
#> 688  -1.508654885      0
#> 689  -0.836233703      0
#> 690  -0.277982498      0
#> 691   0.674373564      0
#> 692   0.798891685      0
#> 693  -1.452542001      0
#> 694  -0.148207017      0
#> 695   1.377511455      0
#> 696   0.210894029      0
#> 697  -0.354867259      0
#> 698   1.053535653      0
#> 699  -0.718522281      0
#> 700  -1.203897444      0
#> 701   1.014391892      0
#> 702  -0.294110553      0
#> 703   0.260019963      0
#> 704   0.600721937      0
#> 705  -0.500510037      0
#> 706  -1.069419515      0
#> 707  -0.491337644      0
#> 708  -0.109480143      0
#> 709  -1.076747980      0
#> 710  -0.347630276      0
#> 711   0.594075718      0
#> 712   0.583149809      0
#> 713  -0.375525725      0
#> 714  -1.543880145      0
#> 715   0.106940178      0
#> 716  -1.476792466      0
#> 717   0.149505233      0
#> 718  -0.310649932      0
#> 719  -0.692291122      0
#> 720  -1.618430208      0
#> 721   0.428112193      0
#> 722  -1.454412966      0
#> 723  -0.775559848      0
#> 724   0.233949418      0
#> 725   0.807736879      0
#> 726   0.116464436      0
#> 727  -0.120613579      0
#> 728   0.705864302      0
#> 729   1.504809055      0
#> 730   0.935068323      0
#> 731   0.452495973      0
#> 732   0.846050088      0
#> 733   0.950833305      0
#> 734  -0.941191464      0
#> 735  -0.533471435      0
#> 736   0.494051728      0
#> 737   0.040704319      0
#> 738  -0.172007430      0
#> 739  -0.786510536      0
#> 740  -0.971230034      0
#> 741   1.056662064      0
#> 742   0.459789571      0
#> 743   0.379364419      0
#> 744   1.342606648      0
#> 745   0.785948121      0
#> 746   1.666289340      0
#> 747  -0.230683802      0
#> 748   1.316463090      0
#> 749   0.648972149      0
#> 750   0.171130110      0
#> 751  -1.388867877      0
#> 752   0.950027249      0
#> 753   0.349220887      0
#> 754   0.275134094      0
#> 755   1.335513973      0
#> 756  -0.195390380      0
#> 757  -1.298889378      0
#> 758  -1.014033379      0
#> 759   0.477390014      0
#> 760  -1.301268237      0
#> 761  -0.346411471      0
#> 762  -1.165332694      0
#> 763  -0.616357152      0
#> 764  -1.018720111      0
#> 765  -0.287812491      0
#> 766  -0.448495617      0
#> 767   0.247506765      0
#> 768  -1.647736318      0
#> 769  -0.143671634      0
#> 770  -1.087209359      0
#> 771  -1.835812734      0
#> 772  -0.074946046      0
#> 773  -0.564489419      0
#> 774   0.524591153      0
#> 775   0.356188816      0
#> 776   0.068642617      0
#> 777  -0.490036503      0
#> 778   0.300539081      0
#> 779  -1.025220214      0
#> 780  -0.662626007      0
#> 781   0.081213386      0
#> 782   1.032572756      0
#> 783  -1.190535245      0
#> 784  -0.922481551      0
#> 785  -0.062296787      0
#> 786   0.799211937      0
#> 787  -0.386445279      0
#> 788   0.524259096      0
#> 789  -1.927807523      0
#> 790   1.476019569      0
#> 791   0.219925687      0
#> 792   0.904469505      0
#> 793  -0.176426047      0
#> 794   0.074822875      0
#> 795  -0.468309921      0
#> 796   1.646627788      0
#> 797   0.319088023      0
#> 798  -0.038252024      0
#> 799  -1.217145661      0
#> 800  -0.996112107      0
#> 801  -0.004291643      0
#> 802  -1.152695133      0
#> 803  -2.030597034      0
#> 804   0.616104268      0
#> 805   1.327536314      0
#> 806  -0.763130594      0
#> 807  -0.347225151      0
#> 808   0.294087821      0
#> 809  -0.219753073      0
#> 810  -0.667370027      0
#> 811  -0.065990378      0
#> 812   0.397826392      0
#> 813   2.102030220      0
#> 814  -0.918991413      0
#> 815  -0.413135533      0
#> 816  -1.394918545      0
#> 817   1.152823778      0
#> 818   0.103536262      0
#> 819   0.886736172      0
#> 820  -0.782190291      0
#> 821  -0.587284503      0
#> 822  -0.687428435      0
#> 823  -0.710870135      0
#> 824  -1.320447939      0
#> 825   1.561187193      0
#> 826   1.369431689      0
#> 827   1.073911944      0
#> 828  -0.738256887      0
#> 829   0.428990948      0
#> 830   0.660890370      0
#> 831  -0.467926015      0
#> 832  -0.938905417      0
#> 833   0.131650408      0
#> 834  -0.316109122      0
#> 835  -0.394591746      0
#> 836  -0.631590640      0
#> 837  -0.666869190      0
#> 838   0.669795787      0
#> 839  -0.482829577      0
#> 840   1.470328556      0
#> 841  -0.069225912      0
#> 842   1.140400387      0
#> 843   0.031496716      0
#> 844   0.412774947      0
#> 845  -0.156573060      0
#> 846  -0.433764161      0
#> 847   0.969171237      0
#> 848   0.960195855      0
#> 849   0.366516443      0
#> 850  -0.539589122      0
#> 851  -1.096890832      0
#> 852   0.547871996      0
#> 853  -1.195698467      0
#> 854  -0.179813793      0
#> 855  -1.258840668      0
#> 856  -0.855744794      0
#> 857   0.357320002      0
#> 858  -1.195989775      0
#> 859  -0.420966146      0
#> 860  -0.548704835      0
#> 861   0.904157737      0
#> 862  -0.354282266      0
#> 863   0.211048325      0
#> 864   2.070237190      0
#> 865   0.385191902      0
#> 866  -1.444773634      0
#> 867   0.939237210      0
#> 868   0.046403352      0
#> 869  -0.596979577      0
#> 870   0.305425866      0
#> 871  -1.125592741      0
#> 872  -0.387360414      0
#> 873  -0.582110676      0
#> 874   1.987920793      0
#> 875   0.374808660      0
#> 876  -1.246958463      0
#> 877   1.362474663      0
#> 878   1.649545871      0
#> 879   0.013684537      0
#> 880  -0.411595486      0
#> 881  -0.169424719      0
#> 882   0.016547723      0
#> 883  -0.134222765      0
#> 884  -0.862178547      0
#> 885   0.098793288      0
#> 886  -1.034265929      0
#> 887  -1.250529774      0
#> 888  -0.720524878      0
#> 889  -0.255577560      0
#> 890  -0.831111531      0
#> 891  -1.339573340      0
#> 892  -0.970418678      0
#> 893  -0.865961980      0
#> 894  -1.270484006      0
#> 895  -0.990178077      0
#> 896  -0.333516427      0
#> 897   0.644041175      0
#> 898  -0.264785340      0
#> 899  -0.448314550      0
#> 900  -2.190433469      0
#> 901   0.273075244      0
#> 902  -0.400915113      0
#> 903  -0.759693101      0
#> 904  -1.467635317      0
#> 905   0.374532761      0
#> 906   0.527037032      0
#> 907   1.242680762      0
#> 908   0.749428083      0
#> 909   2.061052694      1
#> 910  -0.848785485      1
#> 911  -0.353152558      1
#> 912   0.734404205      1
#> 913  -1.290154687      1
#> 914  -0.338943744      1
#> 915   1.847276327      1
#> 916   0.557979256      1
#> 917   1.469964394      1
#> 918   1.101297322      1
#> 919   2.029096083      1
#> 920   3.118010230      1
#> 921  -0.182888107      1
#> 922  -0.523738237      1
#> 923   0.211419650      1
#> 924   1.585536105      1
#> 925   1.208566057      1
#> 926   0.598004479      1
#> 927   0.876955345      1
#> 928   2.039106196      1
#> 929   0.001824054      1
#> 930  -0.039616067      1
#> 931   0.461902934      1
#> 932   1.065135532      1
#> 933   1.983530447      1
#> 934  -0.132946812      1
#> 935   0.004485799      1
#> 936   0.492554982      1
#> 937   0.326540723      1
#> 938  -0.061446681      1
#> 939   0.920540766      1
#> 940   1.919412035      1
#> 941   1.518958923      1
#> 942   1.768323186      1
#> 943  -1.223238366      1
#> 944   0.282061999      1
#> 945   1.537812815      1
#> 946   1.039508152      1
#> 947   1.612644590      1
#> 948   0.213560070      1
#> 949   0.197063485      1
#> 950   1.491357439      1
#> 951   0.541801995      1
#> 952   1.141110498      1
#> 953  -2.003949956      1
#> 954  -0.500736422      1
#> 955   2.098983387      1
#> 956   1.369152375      1
#> 957   1.785595612      1
#> 958   1.070445881      1
#> 959   0.804648632      1
#> 960  -0.351198206      1
#> 961   2.865922704      1
#> 962   0.272761510      1
#> 963   1.227361863      1
#> 964   0.383717342      1
#> 965   0.569302157      1
#> 966   0.418783830      1
#> 967   0.945508717      1
#> 968  -0.573947027      1
#> 969   0.896766615      1
#> 970   1.720134013      1
#> 971   1.526462957      1
#> 972   0.700589265      1
#> 973   2.247380763      1
#> 974   1.117660988      1
#> 975  -1.142753818      1
#> 976   0.047441426      1
#> 977   1.164062409      1
#> 978   2.065394212      1
#> 979   1.749137666      1
#> 980   2.364515054      1
#> 981   0.070118760      1
#> 982   1.036971879      1
#> 983  -0.325105364      1
#> 984   1.374002376      1
#> 985   1.338220937      1
#> 986   2.038831947      1
#> 987   1.007958497      1
#> 988   2.154596564      1
#> 989  -0.041739488      1
#> 990  -0.934742935      1
#> 991   1.948302902      1
#> 992   1.969980080      1
#> 993   0.071537690      1
#> 994  -1.186363865      1
#> 995   0.021470401      1
#> 996  -0.156033025      1
#> 997   0.818402220      1
#> 998  -0.409316745      1
#> 999   1.738918050      1
#> 1000  1.673812264      1