Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -1.4769830472      0
#> 2    -0.2436109310      0
#> 3    -0.2438390619      0
#> 4     0.6049173659      0
#> 5     1.8564597272      0
#> 6    -1.1125727670      0
#> 7     0.3543201580      0
#> 8    -0.3362379149      0
#> 9    -0.4477481407      0
#> 10   -1.2547769508      0
#> 11    0.3304678295      0
#> 12    1.8542706054      0
#> 13   -1.2837133309      0
#> 14   -0.2046359218      0
#> 15   -1.0338297047      0
#> 16    0.3349519966      0
#> 17    0.7611672024      0
#> 18    0.1488898446      0
#> 19    0.0949416655      0
#> 20   -0.4636972831      0
#> 21    1.9711753032      0
#> 22    1.4317645763      0
#> 23    0.7422171877      0
#> 24    0.3687763874      0
#> 25    1.7330788725      0
#> 26    0.3469338217      0
#> 27    0.2015212265      0
#> 28   -0.6102574793      0
#> 29    1.5126130774      0
#> 30   -0.6445679388      0
#> 31   -0.0199391753      0
#> 32   -0.2832905273      0
#> 33    2.3691269156      0
#> 34   -1.7269612766      0
#> 35    1.0657899143      0
#> 36    0.8315706949      0
#> 37   -1.3531776728      0
#> 38   -1.3225660441      0
#> 39    0.7482785486      0
#> 40    2.3634955711      0
#> 41    0.5168380429      0
#> 42    0.7290577931      0
#> 43   -0.8907006730      0
#> 44    0.8603558937      0
#> 45   -0.4556191533      0
#> 46    1.6835598412      0
#> 47   -0.4868938409      0
#> 48   -2.0104646399      0
#> 49    2.1653748935      0
#> 50    0.3160666060      0
#> 51   -0.5716137466      0
#> 52   -1.4306397510      0
#> 53    0.8000190918      0
#> 54   -0.4459222609      0
#> 55   -0.5633500703      0
#> 56   -1.0393273288      0
#> 57    0.8899922362      0
#> 58    1.0134457711      0
#> 59    1.1481097898      0
#> 60    0.0024474873      0
#> 61    0.4419536867      0
#> 62   -0.1485233897      0
#> 63   -0.3598658755      0
#> 64    0.6286122277      0
#> 65    0.5165183576      0
#> 66    0.1851305959      0
#> 67   -0.5547628451      0
#> 68   -1.6051643756      0
#> 69   -0.3323774073      0
#> 70    0.2119428720      0
#> 71   -0.3746011075      0
#> 72   -0.1297437918      0
#> 73    0.8984137667      0
#> 74    0.3640685317      0
#> 75   -2.2350874138      0
#> 76    2.6088130212      0
#> 77    1.3787488697      0
#> 78   -0.2400685808      0
#> 79    1.0757241743      0
#> 80   -0.6992721711      0
#> 81    0.2954147202      0
#> 82    0.3106180224      0
#> 83    0.0516754433      0
#> 84    0.7986163288      0
#> 85   -0.2429013072      0
#> 86   -0.0648667450      0
#> 87    0.3759241851      0
#> 88    0.8370660640      0
#> 89    1.1264143227      0
#> 90    0.4193756666      0
#> 91   -2.0670013875      0
#> 92   -1.4565246253      0
#> 93    0.7133780813      0
#> 94   -2.3193379367      0
#> 95   -0.6723303047      0
#> 96    0.1686677894      0
#> 97    0.8107479850      0
#> 98    2.6410399949      0
#> 99    0.4301965999      0
#> 100   0.4385766844      0
#> 101  -0.5327131455      0
#> 102   0.5528188257      0
#> 103  -1.7686763470      0
#> 104  -0.2894108192      0
#> 105   1.0477372282      0
#> 106  -0.7144974633      0
#> 107  -0.1355259871      0
#> 108  -0.1615044528      0
#> 109   0.9802630178      0
#> 110  -0.7662198379      0
#> 111   0.0141551509      0
#> 112   1.0737993447      0
#> 113  -0.3366985406      0
#> 114   0.8521655361      0
#> 115   0.6716979152      0
#> 116   1.5973950148      0
#> 117   1.6640895834      0
#> 118   0.0505763933      0
#> 119   0.3383307540      0
#> 120   0.4446579954      0
#> 121  -1.9656330620      0
#> 122  -0.5807777651      0
#> 123  -0.2103303520      0
#> 124  -1.1884456522      0
#> 125  -0.9817140975      0
#> 126   0.6840360009      0
#> 127  -1.0120211960      0
#> 128   0.0401533326      0
#> 129  -0.5304036199      0
#> 130  -0.4855607064      0
#> 131   0.3526297912      0
#> 132   0.1784982307      0
#> 133  -0.6742622436      0
#> 134   0.5264165662      0
#> 135   1.2120693618      0
#> 136  -1.6665942012      0
#> 137   0.3101494371      0
#> 138  -2.9445970027      0
#> 139   0.0201217807      0
#> 140  -0.3182176871      0
#> 141  -1.8644035550      0
#> 142   0.2099290841      0
#> 143   1.0486622355      0
#> 144  -0.4438647748      0
#> 145  -0.9117868189      0
#> 146   0.9414489397      0
#> 147  -0.7231857738      0
#> 148  -0.2740350118      0
#> 149   1.5129592734      0
#> 150  -0.1725840008      0
#> 151   0.1515410127      0
#> 152   0.5471656207      0
#> 153  -1.1048404173      0
#> 154   2.0091279837      0
#> 155   1.2693439946      0
#> 156  -1.5921655574      0
#> 157   0.1545114507      0
#> 158   1.8658577273      0
#> 159  -0.8254404058      0
#> 160   0.2878250440      0
#> 161   0.5786391237      0
#> 162  -0.9125706176      0
#> 163  -0.4307933522      0
#> 164  -0.4566614551      0
#> 165   1.0093749588      0
#> 166   0.7701760721      0
#> 167  -0.0855799395      0
#> 168   0.0554643916      0
#> 169   1.9691170983      0
#> 170   0.0389185943      0
#> 171  -0.3409241444      0
#> 172  -0.9148961803      0
#> 173  -0.7752692185      0
#> 174   1.2429372810      0
#> 175  -0.0899934589      0
#> 176  -0.0759481303      0
#> 177   2.1642041015      0
#> 178   0.6441144939      0
#> 179   0.6352974671      0
#> 180   1.0651010622      0
#> 181  -1.5043979353      0
#> 182   0.2479230488      0
#> 183   0.7860221466      0
#> 184   0.1161288700      0
#> 185  -0.1513196760      0
#> 186  -0.5775270567      0
#> 187   1.0551440913      0
#> 188   1.6224592600      0
#> 189  -0.5897401250      0
#> 190  -0.2182520782      0
#> 191  -0.4892739316      0
#> 192   1.6000483434      0
#> 193   2.2113549213      0
#> 194  -0.4211787803      0
#> 195   1.6713850664      0
#> 196   0.5550945293      0
#> 197   0.1611906078      0
#> 198   0.0389756109      0
#> 199   1.3631461500      0
#> 200   0.7420111729      0
#> 201   0.1930255719      0
#> 202   0.2187254981      0
#> 203  -0.0981139531      0
#> 204   0.1241933375      0
#> 205  -0.7205875016      0
#> 206   1.1820720189      0
#> 207  -1.2061969673      0
#> 208   0.7931488621      0
#> 209  -0.2149181920      0
#> 210   1.4180885881      0
#> 211  -0.0524312273      0
#> 212  -0.2852687184      0
#> 213  -2.1805452569      0
#> 214   0.8805838941      0
#> 215  -0.2657887813      0
#> 216  -2.5994905514      0
#> 217   0.8882263406      0
#> 218   0.0928948580      0
#> 219   0.7738565293      0
#> 220  -2.4206741154      0
#> 221   0.1644847545      0
#> 222   1.0953031246      0
#> 223   0.1849000736      0
#> 224  -0.1116603915      0
#> 225   0.5145742924      0
#> 226   1.4023534203      0
#> 227  -1.0779005736      0
#> 228   1.0425619946      0
#> 229  -1.4994340045      0
#> 230  -0.9936138445      0
#> 231   0.4126050083      0
#> 232   1.8022456628      0
#> 233  -0.3230768220      0
#> 234   1.0427393796      0
#> 235  -1.6925827482      0
#> 236   0.6976860869      0
#> 237   1.0160205810      0
#> 238   0.0574094657      0
#> 239  -1.6272067674      0
#> 240  -1.5205380836      0
#> 241   0.9329617619      0
#> 242   0.1702186291      0
#> 243   0.3180950380      0
#> 244  -0.2294892464      0
#> 245   0.2405899275      0
#> 246   0.1888146692      0
#> 247   2.4761134656      0
#> 248   1.2893608843      0
#> 249  -1.8056576203      0
#> 250  -1.1398427176      0
#> 251   1.0825342401      0
#> 252  -0.2488932174      0
#> 253  -0.6900871962      0
#> 254   0.2431027469      0
#> 255  -1.8041989093      0
#> 256  -0.3351558481      0
#> 257   0.4521623464      0
#> 258   1.6342504226      0
#> 259  -0.5635640502      0
#> 260   0.4988457075      0
#> 261   0.2622807826      0
#> 262  -1.5133690526      0
#> 263  -0.7023619743      0
#> 264   0.2212107838      0
#> 265   0.6982189052      0
#> 266  -0.6743065542      0
#> 267   0.1384664662      0
#> 268  -2.8420010747      0
#> 269   0.2156206420      0
#> 270   0.4077205825      0
#> 271  -1.0635461642      0
#> 272   0.3649225133      0
#> 273   1.2920637333      0
#> 274  -0.0200639943      0
#> 275  -0.8609268493      0
#> 276   0.7210645203      0
#> 277   0.2533655430      0
#> 278   1.3240788170      0
#> 279   0.0819477591      0
#> 280  -0.3574877352      0
#> 281   0.8415860391      0
#> 282   0.6053264903      0
#> 283  -0.1960400029      0
#> 284  -0.6311546305      0
#> 285  -0.3853490599      0
#> 286   1.6496427189      0
#> 287   1.6945047573      0
#> 288  -1.3097849183      0
#> 289   0.8210590764      0
#> 290  -1.0084167959      0
#> 291  -1.9338939563      0
#> 292  -0.5386972962      0
#> 293  -1.4223658204      0
#> 294  -0.4724965284      0
#> 295   0.4943865504      0
#> 296   2.0024746761      0
#> 297  -0.0865645084      0
#> 298  -1.2809240832      0
#> 299   0.5619619275      0
#> 300  -1.3837961638      0
#> 301  -0.8067218226      0
#> 302  -1.4329873308      0
#> 303  -0.2991816338      0
#> 304  -0.7692476280      0
#> 305  -0.4571212927      0
#> 306  -0.3595726488      0
#> 307   0.1830964433      0
#> 308  -0.6919480419      0
#> 309  -0.2971963657      0
#> 310   0.2866396122      0
#> 311  -1.4648121153      0
#> 312  -1.0213084760      0
#> 313  -0.7582406583      0
#> 314  -1.1046081004      0
#> 315  -0.2742243889      0
#> 316  -0.2828360159      0
#> 317  -0.3943765004      0
#> 318  -0.0311962218      0
#> 319  -1.2418437809      0
#> 320   0.2672379170      0
#> 321  -0.3255676999      0
#> 322   1.3383235855      0
#> 323   0.0144988120      0
#> 324   0.0993648512      0
#> 325  -0.9379012705      0
#> 326  -0.1204227274      0
#> 327  -1.5034879701      0
#> 328  -0.7935786091      0
#> 329   0.0723863263      0
#> 330  -2.2626183570      0
#> 331  -1.3213932752      0
#> 332  -1.3644672539      0
#> 333  -0.7803375381      0
#> 334   0.8067803508      0
#> 335   0.2765122699      0
#> 336   0.6016017087      0
#> 337  -0.6079372375      0
#> 338   0.4884853251      0
#> 339   1.0447612088      0
#> 340   0.1425936078      0
#> 341  -0.3541211448      0
#> 342  -0.6558646238      0
#> 343  -1.6468892732      0
#> 344  -0.9835391445      0
#> 345  -0.1096157164      0
#> 346   1.7543523995      0
#> 347  -0.5338319160      0
#> 348  -0.5795443526      0
#> 349   0.8659918525      0
#> 350   0.2570046885      0
#> 351   0.9384196785      0
#> 352  -2.0710672962      0
#> 353  -0.0344386058      0
#> 354  -0.1568059218      0
#> 355  -1.3086767702      0
#> 356   0.2111486770      0
#> 357  -0.6227551720      0
#> 358   0.7749001654      0
#> 359   0.3730428901      0
#> 360   0.0423915064      0
#> 361  -0.3727787427      0
#> 362   0.9266974217      0
#> 363  -0.4201627709      0
#> 364  -0.4438528804      0
#> 365  -0.1219039432      0
#> 366   0.3401549778      0
#> 367  -2.1412376802      0
#> 368   2.1607844734      0
#> 369  -0.8105313655      0
#> 370  -0.0852931250      0
#> 371  -2.8007759650      0
#> 372  -0.9006674272      0
#> 373   0.8847128935      0
#> 374  -1.0295758897      0
#> 375   1.2178790987      0
#> 376   1.3146385923      0
#> 377  -0.5083272080      0
#> 378   0.4369194381      0
#> 379   1.1819837516      0
#> 380  -0.2574994373      0
#> 381  -0.6509967415      0
#> 382  -0.7415714842      0
#> 383   0.9184221625      0
#> 384  -0.6304657915      0
#> 385  -2.1118500949      0
#> 386   0.8105368438      0
#> 387  -0.5435416619      0
#> 388   0.5125034174      0
#> 389  -0.8305451652      0
#> 390   0.2295282988      0
#> 391   0.9471709472      0
#> 392   1.0063069234      0
#> 393  -0.0330716016      0
#> 394  -0.8188639033      0
#> 395  -1.0431480202      0
#> 396   0.9955695689      0
#> 397  -0.4714774565      0
#> 398   0.4445749283      0
#> 399  -0.1618818527      0
#> 400   1.0355783912      0
#> 401  -0.5038211136      0
#> 402   0.0499481641      0
#> 403   0.1402596154      0
#> 404  -2.3239997728      0
#> 405  -0.5059290286      0
#> 406  -0.4843245738      0
#> 407   0.2158180240      0
#> 408  -0.9481824139      0
#> 409  -0.0338876456      0
#> 410  -0.5866489269      0
#> 411  -0.8771418618      0
#> 412   0.0735966757      0
#> 413   0.0335548420      0
#> 414  -0.4971636239      0
#> 415  -0.5012909418      0
#> 416  -1.4695143266      0
#> 417  -0.4479691986      0
#> 418   0.1659769401      0
#> 419   0.4753367712      0
#> 420  -0.4210162444      0
#> 421   0.7391304897      0
#> 422  -1.2908839267      0
#> 423   0.1707971384      0
#> 424  -0.8181039869      0
#> 425   0.7301025863      0
#> 426  -2.4080435650      0
#> 427  -0.1571299104      0
#> 428   1.1906246438      0
#> 429  -3.2955213698      0
#> 430  -0.0005429189      0
#> 431   0.5295580423      0
#> 432  -1.4116722936      0
#> 433  -1.8533601768      0
#> 434   0.3899600056      0
#> 435   0.4510833237      0
#> 436   0.7657383960      0
#> 437   1.0777214959      0
#> 438   0.3809277268      0
#> 439  -0.1639245215      0
#> 440   1.1272170866      0
#> 441  -1.2526014068      0
#> 442  -1.3931818821      0
#> 443  -0.3107468270      0
#> 444   0.5360187201      0
#> 445  -0.4797398392      0
#> 446  -0.5175495211      0
#> 447   0.7832193549      0
#> 448  -1.0441985748      0
#> 449   0.5171035525      0
#> 450   0.1470990653      0
#> 451  -1.0435832080      0
#> 452  -0.3291298154      0
#> 453  -1.1405281747      0
#> 454   1.0030200523      0
#> 455   0.2699772275      0
#> 456   0.6532612474      0
#> 457  -0.3940749355      0
#> 458  -1.1357602318      0
#> 459   0.4574165417      0
#> 460  -0.6374304488      0
#> 461   0.1017818135      0
#> 462  -0.1987167128      0
#> 463   0.8063257112      0
#> 464   0.2353624172      0
#> 465  -0.0508118119      0
#> 466  -0.2790863601      0
#> 467  -0.2318809919      0
#> 468   0.2907949534      0
#> 469  -1.0223808600      0
#> 470  -1.4873538593      0
#> 471   0.8816497766      0
#> 472  -0.8019316796      0
#> 473   2.1937512198      0
#> 474  -1.0405887015      0
#> 475   1.1365043282      0
#> 476  -0.6321863799      0
#> 477   0.4185951771      0
#> 478  -0.0822204275      0
#> 479   1.4462859438      0
#> 480   0.6815390093      0
#> 481   0.2024072511      0
#> 482  -1.2738839730      0
#> 483   1.4696818640      0
#> 484   0.2438110322      0
#> 485   0.3902911152      0
#> 486  -1.1582256875      0
#> 487  -0.1588267521      0
#> 488  -1.7858198674      0
#> 489   0.9083565365      0
#> 490   0.0477665204      0
#> 491  -1.0503985514      0
#> 492  -1.3435198407      0
#> 493  -1.8043769869      0
#> 494  -0.2536868723      0
#> 495  -0.9758741841      0
#> 496  -0.0533771098      0
#> 497  -0.3242341068      0
#> 498  -0.1899943035      0
#> 499  -0.0439252768      0
#> 500  -0.3875935470      0
#> 501   1.3442574244      0
#> 502  -0.2343311399      0
#> 503   1.6888826838      0
#> 504  -2.2109832812      0
#> 505  -1.4866630863      0
#> 506  -0.8484286507      0
#> 507  -0.7302388238      0
#> 508   0.7321689058      0
#> 509  -0.5375423288      0
#> 510  -0.8573293165      0
#> 511  -0.3595178793      0
#> 512  -1.7449006332      0
#> 513  -1.0133397184      0
#> 514   0.1302046004      0
#> 515  -0.2373105858      0
#> 516  -1.3140702438      0
#> 517  -0.6035547645      0
#> 518  -1.2563604224      0
#> 519   1.0571848356      0
#> 520   1.8133848837      0
#> 521   0.2841221322      0
#> 522   0.1701447192      0
#> 523  -0.5315501334      0
#> 524  -0.0622092116      0
#> 525   1.0944114674      0
#> 526   1.6562360443      0
#> 527  -1.0277713125      0
#> 528  -0.3227619474      0
#> 529  -0.3769570142      0
#> 530   0.0378173973      0
#> 531   0.7978155918      0
#> 532  -0.3788935209      0
#> 533   0.2572421350      0
#> 534   1.4422534408      0
#> 535   1.0153727595      0
#> 536  -0.0904689152      0
#> 537  -0.6260350403      0
#> 538   0.7355024263      0
#> 539  -0.2592830862      0
#> 540   1.0775703061      0
#> 541  -1.4493368236      0
#> 542   0.6379496386      0
#> 543   0.3821563235      0
#> 544   1.2649259473      0
#> 545   0.0117630584      0
#> 546  -1.2425885633      0
#> 547  -0.8979139898      0
#> 548  -0.2936312141      0
#> 549  -1.8011397378      0
#> 550  -0.0016505286      0
#> 551  -1.2568063892      0
#> 552   0.2465620373      0
#> 553  -1.7542476290      0
#> 554   0.0570589645      0
#> 555  -1.0915218588      0
#> 556   0.8722494022      0
#> 557  -0.1164141538      0
#> 558  -0.1107643173      0
#> 559   1.0998991750      0
#> 560   0.0594140030      0
#> 561   0.9382252265      0
#> 562   0.4160914144      0
#> 563  -0.9177993657      0
#> 564   1.1500314254      0
#> 565  -0.3526031309      0
#> 566  -0.3139021456      0
#> 567  -1.4520131118      0
#> 568   0.6976928291      0
#> 569  -0.4394281845      0
#> 570  -1.0449991378      0
#> 571   0.1057178248      0
#> 572   0.7051918740      0
#> 573   2.5015822124      0
#> 574   0.6763509397      0
#> 575  -0.4491013521      0
#> 576   0.2389098577      0
#> 577  -2.6721369287      0
#> 578  -1.8454939416      0
#> 579   0.6717594248      0
#> 580  -0.9904949907      0
#> 581   0.7211511004      0
#> 582  -0.3405640397      0
#> 583   1.3441878300      0
#> 584  -0.6895835742      0
#> 585   0.1043214490      0
#> 586  -0.9286605990      0
#> 587   1.1219532922      0
#> 588  -0.9281534704      0
#> 589   1.2824538567      0
#> 590  -0.0521053538      0
#> 591  -0.7798224478      0
#> 592   1.6305044157      0
#> 593  -0.6469269892      0
#> 594  -0.0275854189      0
#> 595   0.8677627530      0
#> 596   1.4919103424      0
#> 597   1.7871188700      0
#> 598   0.3726809108      0
#> 599   0.5601421429      0
#> 600  -0.2963384785      0
#> 601   0.8927861555      0
#> 602   1.0695644593      0
#> 603  -0.0182044703      0
#> 604  -0.9512114256      0
#> 605   1.3104753097      0
#> 606  -0.5119501584      0
#> 607  -1.6152671786      0
#> 608   0.4001167692      0
#> 609   1.4574585302      0
#> 610  -1.4239415668      0
#> 611  -0.7137245064      0
#> 612   1.0612051167      0
#> 613   0.7020298477      0
#> 614   0.5040743779      0
#> 615  -0.4492760196      0
#> 616  -0.3389133423      0
#> 617  -0.5905165205      0
#> 618  -1.1039454604      0
#> 619   1.4760717928      0
#> 620   0.6476080689      0
#> 621  -1.4991704609      0
#> 622  -2.6454901139      0
#> 623  -0.1656490499      0
#> 624   0.1231245606      0
#> 625   1.0075089912      0
#> 626  -1.6385488714      0
#> 627   1.1822080369      0
#> 628  -1.1996399134      0
#> 629   0.0003223366      0
#> 630   1.9959322725      0
#> 631   0.9294543024      0
#> 632   0.1361458848      0
#> 633  -1.8889308715      0
#> 634  -0.5172228664      0
#> 635   0.3864288406      0
#> 636   0.5430796103      0
#> 637   1.0118797705      0
#> 638  -0.2698340204      0
#> 639  -1.4782308404      0
#> 640  -1.9756995426      0
#> 641  -0.0664080040      0
#> 642   1.1796629137      0
#> 643  -0.9920244040      0
#> 644   0.2961932693      0
#> 645  -0.9170654066      0
#> 646  -2.5525040023      0
#> 647   0.8060238160      0
#> 648  -0.0571413838      0
#> 649  -0.6019677183      0
#> 650  -1.8404246163      0
#> 651  -0.0602528265      0
#> 652   2.1252393771      0
#> 653  -0.9361461550      0
#> 654   0.6219467217      0
#> 655   0.6082890306      0
#> 656  -1.4083453104      0
#> 657   0.4867181748      0
#> 658  -2.0207734825      0
#> 659  -0.6562994666      0
#> 660  -0.7086413699      0
#> 661   0.5686133058      0
#> 662  -1.6127557944      0
#> 663   0.2556806600      0
#> 664   1.1242604056      0
#> 665   1.1864669248      0
#> 666   0.3534124155      0
#> 667   1.8793916942      0
#> 668   0.9639556182      0
#> 669   0.1630418350      0
#> 670   0.5709060250      0
#> 671  -2.5866918440      0
#> 672  -0.6379006921      0
#> 673  -0.5570987133      0
#> 674   0.9599209228      0
#> 675  -0.6088854160      0
#> 676   0.7863535507      0
#> 677   0.6258431481      0
#> 678  -0.3024684376      0
#> 679  -1.2691024586      0
#> 680  -1.4619362794      0
#> 681  -0.3010805443      0
#> 682   1.3658872653      0
#> 683   1.0390761169      0
#> 684  -0.9391587970      0
#> 685  -1.8556979400      0
#> 686   0.4321777222      0
#> 687  -1.3795296367      0
#> 688   1.3880278717      0
#> 689   0.2774328393      0
#> 690   0.1964376651      0
#> 691   2.4491563439      0
#> 692  -0.4140208616      0
#> 693   0.2500877734      0
#> 694   0.4823413111      0
#> 695   0.7119095448      0
#> 696   0.1708773117      0
#> 697  -0.1487417203      0
#> 698   2.0520670198      0
#> 699   1.3317544281      0
#> 700   0.6633602638      0
#> 701  -0.4004461086      0
#> 702  -0.2077575360      0
#> 703   0.1589683916      0
#> 704   0.0462728717      0
#> 705   0.7729362273      0
#> 706   2.8206636734      0
#> 707   0.4264374891      0
#> 708  -0.6928311595      0
#> 709   0.8087758719      0
#> 710  -1.3951792432      0
#> 711  -0.1564313748      0
#> 712   0.8255041880      0
#> 713   1.9016784475      0
#> 714  -0.5581455582      0
#> 715  -0.2148130964      0
#> 716  -0.0303071113      0
#> 717   0.4976662014      0
#> 718   2.5882597133      0
#> 719   0.0637517830      0
#> 720   0.8654053835      0
#> 721   0.5781255676      0
#> 722  -0.5124113730      0
#> 723  -0.1497123623      0
#> 724  -1.0494568978      0
#> 725   0.7349873961      0
#> 726   0.7308480217      0
#> 727  -0.8013462905      0
#> 728  -0.1210133158      0
#> 729  -0.6390829260      0
#> 730  -0.1575306109      0
#> 731  -0.9724632776      0
#> 732   0.3696635369      0
#> 733   1.5175275678      0
#> 734   1.1257362294      0
#> 735  -0.7388712347      0
#> 736   0.4412964547      0
#> 737  -1.5620154982      0
#> 738  -1.1262200779      0
#> 739   0.5758364626      0
#> 740   0.9372254122      0
#> 741  -0.7353965128      0
#> 742  -1.3620092122      0
#> 743   0.3735215839      0
#> 744   0.6003177334      0
#> 745   0.9031334775      0
#> 746   0.9028196178      0
#> 747  -1.1597916393      0
#> 748   0.3240750015      0
#> 749   1.7289301789      0
#> 750   0.3409451911      0
#> 751   0.2536806428      0
#> 752   1.0044258205      0
#> 753   0.8448537899      0
#> 754  -0.2524647660      0
#> 755   0.9024549652      0
#> 756   0.5361754660      0
#> 757   1.1452875745      0
#> 758  -0.0573494033      0
#> 759  -0.9094580259      0
#> 760   1.3850274528      0
#> 761  -0.0328800288      0
#> 762  -0.4010923454      0
#> 763  -0.3495844466      0
#> 764   0.5567470080      0
#> 765   0.0847122928      0
#> 766   0.5838480970      0
#> 767   0.1218089629      0
#> 768  -0.0102643833      0
#> 769  -0.0126553955      0
#> 770   2.6579887145      0
#> 771  -0.8979222633      0
#> 772  -1.9518834219      0
#> 773   0.1613114511      0
#> 774   0.2768542119      0
#> 775  -1.0059188239      0
#> 776  -0.0649332914      0
#> 777  -0.8917735408      0
#> 778  -0.1205487359      0
#> 779   0.7884789698      0
#> 780  -0.0585685044      0
#> 781  -0.3292212677      0
#> 782   0.3590163227      0
#> 783   1.7503286633      0
#> 784   0.1490880451      0
#> 785   0.1875404047      0
#> 786   1.2237440220      0
#> 787   0.4751372639      0
#> 788  -0.4658788730      0
#> 789   2.1313394833      0
#> 790   1.9576158989      0
#> 791  -0.4514249079      0
#> 792  -0.0984699787      0
#> 793   0.9712219164      0
#> 794  -1.0060565846      0
#> 795   1.2073413271      0
#> 796   0.8502515758      0
#> 797   0.9466287989      0
#> 798  -0.3225072142      0
#> 799   2.2313526284      0
#> 800   0.2072324123      0
#> 801   0.6754758907      0
#> 802  -0.1504846326      0
#> 803  -1.3180779143      0
#> 804   1.0420397156      0
#> 805   0.4460945092      0
#> 806   0.1254280200      0
#> 807   0.0866412377      0
#> 808   0.3788852301      0
#> 809  -0.5302979783      0
#> 810   0.7947453762      0
#> 811  -0.2456260707      0
#> 812   2.4530529788      0
#> 813  -0.1907433962      0
#> 814  -0.3291487069      0
#> 815   0.4751874063      0
#> 816   0.5815536490      0
#> 817   0.1079423618      0
#> 818  -0.0662498255      0
#> 819   0.6603775027      0
#> 820   0.8436902508      0
#> 821  -0.1656366309      0
#> 822  -0.3083198940      0
#> 823  -1.3583554377      0
#> 824   0.0864608591      0
#> 825  -1.7425086048      0
#> 826   0.9261126254      0
#> 827   1.7975307089      0
#> 828  -0.4343723031      0
#> 829  -0.7792143099      0
#> 830   0.4978600493      0
#> 831   0.6931351205      0
#> 832   0.6172998728      0
#> 833  -1.3933945308      0
#> 834  -0.5518408421      0
#> 835   0.7768552120      0
#> 836  -2.6962266008      0
#> 837  -0.5950137228      0
#> 838   0.2958684147      0
#> 839  -0.1133695234      0
#> 840   0.1291471844      0
#> 841   0.5748102802      0
#> 842  -0.3755235345      0
#> 843  -0.5511677141      0
#> 844  -1.2334111921      0
#> 845  -0.1923482725      0
#> 846  -0.7539111135      0
#> 847  -0.0735825678      0
#> 848  -1.2862145618      0
#> 849  -0.9538150696      0
#> 850  -1.4010414356      0
#> 851   0.0081682882      0
#> 852   0.0558603169      0
#> 853  -1.4529502752      0
#> 854   0.5825841817      0
#> 855   0.8151443342      0
#> 856  -0.9604993679      0
#> 857   0.6999891755      0
#> 858   0.3913491162      0
#> 859  -0.2708140726      0
#> 860   0.5115939975      0
#> 861   1.5463720726      0
#> 862   0.1000474751      0
#> 863   0.0971029859      0
#> 864   0.6286398797      0
#> 865   0.9265773648      0
#> 866  -0.6308166001      0
#> 867  -0.9852475247      0
#> 868   0.0125831631      0
#> 869  -0.0857748290      0
#> 870  -0.7548650682      0
#> 871  -1.5198026263      0
#> 872  -0.4771619340      0
#> 873   0.1547745796      0
#> 874  -0.8354198664      0
#> 875  -1.0922365382      0
#> 876  -0.6834772184      0
#> 877  -1.1932483704      0
#> 878   2.0704339281      0
#> 879  -0.4945167013      0
#> 880   0.5430469209      0
#> 881   0.6953005939      0
#> 882  -0.0263163626      0
#> 883  -0.4731007421      0
#> 884   0.1520854436      0
#> 885  -0.8474877060      0
#> 886   1.4767643668      0
#> 887   0.3379451932      0
#> 888   0.4714740932      0
#> 889   0.3085268384      0
#> 890  -1.1944227275      0
#> 891  -0.8475519940      0
#> 892   0.8053632270      0
#> 893   2.2066052248      1
#> 894   1.4341533046      1
#> 895   0.4345698869      1
#> 896  -0.2257112347      1
#> 897  -1.7091740713      1
#> 898   0.2100021830      1
#> 899   1.0515411280      1
#> 900   0.6056127661      1
#> 901   0.8683941683      1
#> 902   0.4351029716      1
#> 903   0.2000626022      1
#> 904   0.3094544803      1
#> 905   0.1006736603      1
#> 906   1.1183850232      1
#> 907   1.3446504433      1
#> 908   0.5493712310      1
#> 909   1.3839698367      1
#> 910   1.3463511906      1
#> 911   1.7814109677      1
#> 912   2.1053886909      1
#> 913  -1.0821396860      1
#> 914   0.9343689724      1
#> 915   1.0258194799      1
#> 916   0.5323384934      1
#> 917   2.0788100432      1
#> 918   1.6450128067      1
#> 919   1.1876460062      1
#> 920   1.3044604689      1
#> 921  -0.2183273322      1
#> 922   1.7247567668      1
#> 923  -0.1412167120      1
#> 924   1.6528223325      1
#> 925  -1.5516235288      1
#> 926   0.8541650039      1
#> 927   2.1800365576      1
#> 928  -0.1940485107      1
#> 929   2.8365515423      1
#> 930   0.6890392425      1
#> 931  -0.3062940900      1
#> 932   1.4950686324      1
#> 933   0.4147993998      1
#> 934   1.4492819725      1
#> 935  -0.6062079926      1
#> 936   0.0940419379      1
#> 937   1.6433072676      1
#> 938   0.2788728470      1
#> 939   0.4766599109      1
#> 940   0.1206888229      1
#> 941  -1.5459878310      1
#> 942   0.0945858857      1
#> 943  -0.6131893692      1
#> 944   1.2143219879      1
#> 945   1.8003453147      1
#> 946   2.9863316491      1
#> 947   1.2330880065      1
#> 948   1.9216081143      1
#> 949   0.5494182481      1
#> 950   0.1888988643      1
#> 951   0.9641797129      1
#> 952  -0.0692704376      1
#> 953  -0.7597912812      1
#> 954   1.1776609950      1
#> 955   1.1000576502      1
#> 956   0.2863230821      1
#> 957   0.6261374798      1
#> 958  -1.0845384164      1
#> 959   1.0242209783      1
#> 960   0.5088182237      1
#> 961   1.3981890920      1
#> 962  -0.2401715849      1
#> 963   1.1403256485      1
#> 964   0.9316477456      1
#> 965   0.6445437306      1
#> 966   1.7408688225      1
#> 967   0.6881389251      1
#> 968   1.5337562674      1
#> 969  -0.2458162989      1
#> 970   0.3015788012      1
#> 971   1.6044490097      1
#> 972   0.1936007786      1
#> 973   0.8740539377      1
#> 974   2.5541194243      1
#> 975   2.9442858706      1
#> 976   0.2979987978      1
#> 977  -0.1600506980      1
#> 978  -0.3837048376      1
#> 979   0.4167072079      1
#> 980  -0.4758525096      1
#> 981   0.7352442773      1
#> 982   1.2291032170      1
#> 983   0.1909332788      1
#> 984  -0.1589275126      1
#> 985   0.1471120546      1
#> 986  -0.8150629859      1
#> 987   2.8650100541      1
#> 988   2.3300148229      1
#> 989   1.6146095161      1
#> 990   0.0483938902      1
#> 991   1.4399276080      1
#> 992   0.2680094757      1
#> 993   0.6110219672      1
#> 994   2.3910808536      1
#> 995   1.9544280561      1
#> 996   0.2328232548      1
#> 997   2.3157812891      1
#> 998  -1.4778438212      1
#> 999   0.9901079369      1
#> 1000 -0.2162335946      1