Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.397466180      0
#> 2     0.810292527      0
#> 3    -2.282284330      0
#> 4     1.590592616      0
#> 5     0.705388277      0
#> 6     0.538165028      0
#> 7     1.681520589      0
#> 8    -0.987154374      0
#> 9    -1.479702036      0
#> 10    0.800419834      0
#> 11    1.884237842      0
#> 12    0.541952505      0
#> 13    0.329181327      0
#> 14   -0.861248647      0
#> 15    1.220928806      0
#> 16    0.135806736      0
#> 17    0.477648012      0
#> 18   -1.125339882      0
#> 19    0.921988444      0
#> 20   -0.534901501      0
#> 21   -0.604403525      0
#> 22    0.342916450      0
#> 23   -0.567087656      0
#> 24   -1.348053634      0
#> 25    0.064035276      0
#> 26    1.052323086      0
#> 27    0.462812784      0
#> 28    0.123183484      0
#> 29    0.912229954      0
#> 30   -0.899491346      0
#> 31    0.439676819      0
#> 32   -0.154557923      0
#> 33    0.589333291      0
#> 34    1.058638533      0
#> 35    0.687652407      0
#> 36    0.543680677      0
#> 37   -0.899854220      0
#> 38    2.126480312      0
#> 39   -0.879841055      0
#> 40   -1.119279880      0
#> 41   -1.218048530      0
#> 42   -1.096697104      0
#> 43    0.333792559      0
#> 44   -0.491053194      0
#> 45    0.564195226      0
#> 46   -1.843436433      0
#> 47   -0.093591009      0
#> 48   -0.042755689      0
#> 49    0.490985047      0
#> 50    0.954912538      0
#> 51   -0.061175142      0
#> 52   -0.934945402      0
#> 53    0.372762061      0
#> 54    0.621449048      0
#> 55   -0.346748311      0
#> 56   -0.361094743      0
#> 57   -1.483278559      0
#> 58   -0.132339620      0
#> 59   -0.632506654      0
#> 60   -0.013368199      0
#> 61    0.144678923      0
#> 62   -0.072441490      0
#> 63    0.579229296      0
#> 64   -0.181541100      0
#> 65   -0.783692224      0
#> 66   -0.082900333      0
#> 67   -0.451143038      0
#> 68    1.712252215      0
#> 69   -0.112273985      0
#> 70   -0.610467839      0
#> 71   -1.383477692      0
#> 72   -2.047350512      0
#> 73    0.017484102      0
#> 74   -1.874549552      0
#> 75   -1.684534747      0
#> 76    0.101947115      0
#> 77    0.372333477      0
#> 78   -1.580719962      0
#> 79   -1.238851390      0
#> 80    1.105782792      0
#> 81    1.497733677      0
#> 82   -0.582416786      0
#> 83   -0.072197029      0
#> 84    0.309910739      0
#> 85   -1.997605460      0
#> 86    1.301101323      0
#> 87    0.708316021      0
#> 88   -0.430450448      0
#> 89   -0.175807855      0
#> 90    0.638288662      0
#> 91   -0.778111543      0
#> 92   -0.404627880      0
#> 93   -0.206458084      0
#> 94   -0.960532828      0
#> 95   -0.142792085      0
#> 96   -0.811662453      0
#> 97   -0.388363515      0
#> 98   -1.422395065      0
#> 99   -1.500645239      0
#> 100   1.621091883      0
#> 101   1.220957065      0
#> 102   2.330280348      0
#> 103   0.629097681      0
#> 104  -0.673360698      0
#> 105   2.810589411      0
#> 106   0.329010756      0
#> 107   1.060864882      0
#> 108   0.216336443      0
#> 109   0.074717952      0
#> 110   0.749026734      0
#> 111   0.258721431      0
#> 112  -1.146905604      0
#> 113  -2.084123209      0
#> 114   1.765244999      0
#> 115  -0.885813335      0
#> 116  -2.163085265      0
#> 117   0.101305960      0
#> 118   1.965460494      0
#> 119   1.423273461      0
#> 120  -1.428585680      0
#> 121   2.183177741      0
#> 122  -0.762247368      0
#> 123  -0.777537238      0
#> 124   0.836151452      0
#> 125   1.105520655      0
#> 126   1.746239009      0
#> 127   0.229470650      0
#> 128  -1.193404456      0
#> 129  -0.863802510      0
#> 130   2.512657944      0
#> 131  -0.920293359      0
#> 132  -1.232646604      0
#> 133   0.564837919      0
#> 134  -0.911089372      0
#> 135   0.609337467      0
#> 136   0.480707019      0
#> 137   0.964604275      0
#> 138  -0.025603245      0
#> 139   1.779979434      0
#> 140  -0.352290960      0
#> 141  -0.685602978      0
#> 142  -0.460340303      0
#> 143  -0.677012151      0
#> 144   0.539207001      0
#> 145  -0.970395443      0
#> 146   1.101465583      0
#> 147   0.723963918      0
#> 148  -0.139617366      0
#> 149  -0.412973063      0
#> 150   0.521968267      0
#> 151  -0.325584781      0
#> 152   0.383600577      0
#> 153   0.371540940      0
#> 154   1.024267927      0
#> 155   0.405332858      0
#> 156  -0.224974339      0
#> 157  -0.737438359      0
#> 158  -0.974525067      0
#> 159  -0.825426206      0
#> 160  -0.645833706      0
#> 161  -1.032162695      0
#> 162  -0.256954284      0
#> 163   0.783837172      0
#> 164   0.810484744      0
#> 165  -0.666267692      0
#> 166   0.030665435      0
#> 167  -1.245336780      0
#> 168   0.736637705      0
#> 169   0.307085010      0
#> 170  -1.113973154      0
#> 171   0.297721475      0
#> 172   0.357322846      0
#> 173   0.683253045      0
#> 174  -1.363287371      0
#> 175  -0.272974142      0
#> 176  -1.358298243      0
#> 177   1.005202910      0
#> 178  -0.140174518      0
#> 179  -1.774574275      0
#> 180   0.037479377      0
#> 181   0.526432638      0
#> 182   0.540490032      0
#> 183  -0.283266372      0
#> 184   2.041607312      0
#> 185   1.087614951      0
#> 186  -0.892773156      0
#> 187   0.493089014      0
#> 188   1.256687411      0
#> 189   1.807614299      0
#> 190   0.366205877      0
#> 191  -1.653414198      0
#> 192  -0.478328911      0
#> 193  -0.580269151      0
#> 194  -1.115510345      0
#> 195  -2.065684588      0
#> 196  -1.396243402      0
#> 197  -1.173242595      0
#> 198   0.144452736      0
#> 199  -0.049676792      0
#> 200   2.377694650      0
#> 201  -0.974498650      0
#> 202  -0.156579000      0
#> 203  -0.213143336      0
#> 204   0.821772267      0
#> 205   0.240063379      0
#> 206   0.833710929      0
#> 207   0.520228784      0
#> 208  -0.068203650      0
#> 209  -0.486719710      0
#> 210  -0.006913069      0
#> 211   0.262341482      0
#> 212   0.024966589      0
#> 213  -0.506665041      0
#> 214  -0.528932245      0
#> 215  -1.559579897      0
#> 216   2.488958157      0
#> 217  -0.984525034      0
#> 218   0.440960279      0
#> 219   0.621007207      0
#> 220  -0.954805740      0
#> 221  -0.091283786      0
#> 222   1.896555115      0
#> 223   1.124988198      0
#> 224  -0.660264643      0
#> 225   0.317911324      0
#> 226   0.231126911      0
#> 227   0.302258221      0
#> 228   1.167624441      0
#> 229  -0.780283171      0
#> 230  -1.222228330      0
#> 231   0.761332833      0
#> 232  -0.806186197      0
#> 233   1.111720207      0
#> 234   0.415097189      0
#> 235   0.612992685      0
#> 236  -0.143405700      0
#> 237  -0.681619494      0
#> 238  -0.033335901      0
#> 239  -0.375872986      0
#> 240  -1.769946906      0
#> 241   2.993521065      0
#> 242   0.027457755      0
#> 243   0.787910212      0
#> 244  -0.469645532      0
#> 245  -2.057968728      0
#> 246   1.040759461      0
#> 247   0.257520272      0
#> 248   0.663680285      0
#> 249  -0.981555256      0
#> 250   0.122832527      0
#> 251  -0.287633332      0
#> 252   0.519076014      0
#> 253  -1.647609555      0
#> 254   1.788736505      0
#> 255   0.734334932      0
#> 256  -0.824999341      0
#> 257   1.477260684      0
#> 258   0.636973797      0
#> 259  -0.863665016      0
#> 260  -0.680036204      0
#> 261   0.129901566      0
#> 262   0.201738194      0
#> 263  -1.772620677      0
#> 264   0.642549472      0
#> 265  -0.245207030      0
#> 266   0.726188486      0
#> 267   0.329638846      0
#> 268   1.571880445      0
#> 269  -0.206437416      0
#> 270   1.208670468      0
#> 271  -2.077099958      0
#> 272  -0.375187095      0
#> 273   1.283308050      0
#> 274   0.058738246      0
#> 275   0.348061597      0
#> 276   0.321304199      0
#> 277   0.104337092      0
#> 278   0.675766027      0
#> 279  -0.086155596      0
#> 280   1.789124227      0
#> 281   0.072635116      0
#> 282  -0.351050173      0
#> 283  -0.043306085      0
#> 284  -1.477428558      0
#> 285   0.219442007      0
#> 286  -0.224920303      0
#> 287  -0.320508871      0
#> 288  -0.289256063      0
#> 289   1.735727328      0
#> 290   0.034317856      0
#> 291  -1.339731506      0
#> 292  -0.359523077      0
#> 293  -0.220940791      0
#> 294  -0.952434054      0
#> 295  -0.332094889      0
#> 296  -0.628984530      0
#> 297  -1.145398489      0
#> 298  -0.674755450      0
#> 299  -1.016140239      0
#> 300   0.912243180      0
#> 301   1.319700195      0
#> 302  -0.214118530      0
#> 303   0.429268579      0
#> 304   0.350783567      0
#> 305   0.260049240      0
#> 306  -0.673517516      0
#> 307  -0.143924757      0
#> 308  -0.026547044      0
#> 309  -0.677115385      0
#> 310  -1.183664275      0
#> 311  -0.029606426      0
#> 312   1.062100286      0
#> 313   1.370867370      0
#> 314  -1.885296566      0
#> 315   0.403637698      0
#> 316  -1.001033257      0
#> 317   1.369307559      0
#> 318   0.726609062      0
#> 319  -0.173588422      0
#> 320  -0.620470396      0
#> 321  -0.804945002      0
#> 322  -0.025902054      0
#> 323  -1.109991502      0
#> 324   1.806950452      0
#> 325  -0.479629119      0
#> 326   0.878972366      0
#> 327   0.283024662      0
#> 328  -0.071267655      0
#> 329   0.097577538      0
#> 330  -0.397450668      0
#> 331  -0.958854563      0
#> 332   1.366900612      0
#> 333   1.304384768      0
#> 334   0.814401297      0
#> 335  -0.293793660      0
#> 336  -0.705066576      0
#> 337   1.057110801      0
#> 338  -2.580415611      0
#> 339  -1.357477697      0
#> 340  -1.226171191      0
#> 341   0.018516742      0
#> 342  -0.164839040      0
#> 343  -0.685057786      0
#> 344   1.549697031      0
#> 345  -1.344038629      0
#> 346   0.029806878      0
#> 347   0.695888219      0
#> 348   0.585200964      0
#> 349  -0.841253485      0
#> 350   0.791713125      0
#> 351  -1.096118316      0
#> 352  -2.699620900      0
#> 353   0.179253757      0
#> 354   1.348223745      0
#> 355   0.612685769      0
#> 356   0.125206622      0
#> 357  -0.960579284      0
#> 358   0.781906742      0
#> 359  -1.808650401      0
#> 360   0.967666140      0
#> 361  -0.626167373      0
#> 362  -0.090268772      0
#> 363   2.621481807      0
#> 364   0.420633899      0
#> 365   1.353801331      0
#> 366   0.130120896      0
#> 367  -1.543394559      0
#> 368  -0.751214286      0
#> 369  -0.715249029      0
#> 370   0.388301068      0
#> 371   0.645641509      0
#> 372   0.244409745      0
#> 373  -1.152421530      0
#> 374   0.545139361      0
#> 375  -0.261003042      0
#> 376   0.292202964      0
#> 377  -0.379529511      0
#> 378  -1.330686192      0
#> 379   1.734656290      0
#> 380  -1.715104289      0
#> 381  -0.521298768      0
#> 382  -0.319865090      0
#> 383  -0.622484589      0
#> 384  -1.359976390      0
#> 385  -0.818796065      0
#> 386   0.706015071      0
#> 387   0.063105496      0
#> 388   1.927367195      0
#> 389  -0.623595643      0
#> 390  -0.720217000      0
#> 391  -0.247056953      0
#> 392   0.856825550      0
#> 393  -0.363330485      0
#> 394  -1.963341431      0
#> 395  -0.980475911      0
#> 396   1.138298986      0
#> 397   0.775679732      0
#> 398   0.375777917      0
#> 399   1.184242577      0
#> 400   0.078280755      0
#> 401  -0.533527674      0
#> 402  -0.128184993      0
#> 403   1.300975859      0
#> 404   0.111783823      0
#> 405   0.636710002      0
#> 406  -0.466334859      0
#> 407  -1.167088383      0
#> 408   1.117923937      0
#> 409   0.477183165      0
#> 410  -0.695154322      0
#> 411  -1.439157448      0
#> 412   0.052574271      0
#> 413  -0.142607033      0
#> 414   0.383886657      0
#> 415   0.706964220      0
#> 416  -0.299331036      0
#> 417   0.844162781      0
#> 418  -0.765685630      0
#> 419   0.911382870      0
#> 420   1.712452509      0
#> 421   1.193990279      0
#> 422   0.299043513      0
#> 423  -0.509477494      0
#> 424   0.296793772      0
#> 425   0.803425647      0
#> 426  -0.496919586      0
#> 427   0.801424808      0
#> 428   1.375979957      0
#> 429   0.721861683      0
#> 430   1.702359084      0
#> 431  -0.288416519      0
#> 432  -0.895427877      0
#> 433  -1.126124362      0
#> 434   1.126192340      0
#> 435  -0.429361629      0
#> 436  -2.384375954      0
#> 437  -0.073260534      0
#> 438   1.388241684      0
#> 439  -0.627697799      0
#> 440  -0.600037339      0
#> 441   0.386694619      0
#> 442  -0.337958477      0
#> 443  -0.768935075      0
#> 444  -0.933751216      0
#> 445   1.795582036      0
#> 446   0.236331739      0
#> 447  -0.211935529      0
#> 448   1.114178228      0
#> 449  -0.676045642      0
#> 450   0.218374706      0
#> 451  -2.278366823      0
#> 452   0.139396481      0
#> 453  -0.462626667      0
#> 454   0.607501016      0
#> 455  -0.556826427      0
#> 456  -0.275618138      0
#> 457  -1.922292633      0
#> 458   0.453467241      0
#> 459   0.591513585      0
#> 460  -1.314158727      0
#> 461  -1.377381081      0
#> 462   0.251934045      0
#> 463   0.576430639      0
#> 464   0.121581559      0
#> 465   1.744808062      0
#> 466  -0.263472291      0
#> 467   0.491397988      0
#> 468  -3.261535965      0
#> 469   0.768140349      0
#> 470   0.868528101      0
#> 471   0.280417572      0
#> 472   1.435826408      0
#> 473  -1.343639128      0
#> 474  -0.561994435      0
#> 475  -0.593095071      0
#> 476  -0.758195992      0
#> 477   0.071027489      0
#> 478   1.132932828      0
#> 479  -1.280247552      0
#> 480   0.771810601      0
#> 481   1.147821577      0
#> 482  -1.063084445      0
#> 483  -0.330314308      0
#> 484  -1.272877019      0
#> 485  -1.846669389      0
#> 486   0.043901346      0
#> 487  -1.205999098      0
#> 488  -0.895933556      0
#> 489   0.637977552      0
#> 490  -0.148201384      0
#> 491   0.547451471      0
#> 492  -0.145128342      0
#> 493  -0.678081646      0
#> 494   0.432991477      0
#> 495  -0.461434070      0
#> 496   1.174317988      0
#> 497  -0.285294562      0
#> 498  -0.107485830      0
#> 499   0.902629398      0
#> 500  -1.015635864      0
#> 501   0.381910694      0
#> 502   0.243260391      0
#> 503   0.421339099      0
#> 504   0.767305286      0
#> 505   0.246472335      0
#> 506   0.798714656      0
#> 507   0.707460091      0
#> 508  -0.608730251      0
#> 509  -0.136538604      0
#> 510  -0.406453764      0
#> 511   0.157266238      0
#> 512   0.769553395      0
#> 513  -1.397348777      0
#> 514   0.616376770      0
#> 515   0.751128427      0
#> 516  -0.373636293      0
#> 517   0.055657204      0
#> 518   1.734807763      0
#> 519  -0.405154343      0
#> 520  -0.083126886      0
#> 521   1.461068719      0
#> 522  -0.404980285      0
#> 523   0.406268194      0
#> 524  -1.167406075      0
#> 525  -0.583763259      0
#> 526   0.430298924      0
#> 527   1.131829465      0
#> 528   0.392523858      0
#> 529  -1.357412301      0
#> 530  -1.425879543      0
#> 531  -0.871399541      0
#> 532   0.135623838      0
#> 533  -1.464532042      0
#> 534  -0.354124911      0
#> 535   0.930240209      0
#> 536   0.233501370      0
#> 537   1.853331812      0
#> 538   0.030901212      0
#> 539  -1.648092888      0
#> 540   0.369416047      0
#> 541  -0.165232475      0
#> 542   1.955835648      0
#> 543   1.528227458      0
#> 544  -0.404464690      0
#> 545  -0.215127315      0
#> 546   0.175908825      0
#> 547   0.062535883      0
#> 548   0.089257268      0
#> 549   1.623270882      0
#> 550   1.351683875      0
#> 551   0.839359871      0
#> 552   0.675496417      0
#> 553   0.567653885      0
#> 554   0.579808129      0
#> 555  -1.200672137      0
#> 556   1.556945663      0
#> 557  -0.434425855      0
#> 558   0.533407284      0
#> 559  -0.221228283      0
#> 560  -0.646108591      0
#> 561   0.771933216      0
#> 562  -0.526191431      0
#> 563   0.504717971      0
#> 564  -1.322920611      0
#> 565  -1.036181786      0
#> 566   1.281922503      0
#> 567   0.286429781      0
#> 568   2.248511794      0
#> 569  -1.226866983      0
#> 570  -1.222370081      0
#> 571  -0.053907868      0
#> 572   0.237402302      0
#> 573   1.251995213      0
#> 574  -0.080179906      0
#> 575   0.590246624      0
#> 576   0.786406642      0
#> 577  -1.893179244      0
#> 578   0.759081799      0
#> 579  -0.812006291      0
#> 580   0.905215337      0
#> 581  -0.378128090      0
#> 582   0.778469304      0
#> 583  -0.315414367      0
#> 584  -0.016434279      0
#> 585  -0.064347217      0
#> 586  -1.031815452      0
#> 587  -0.720062469      0
#> 588   0.553641111      0
#> 589   0.462961542      0
#> 590   0.493490687      0
#> 591  -1.747235334      0
#> 592   0.159331628      0
#> 593   1.137095689      0
#> 594  -0.520180977      0
#> 595  -0.014279210      0
#> 596   0.674193023      0
#> 597  -1.019304210      0
#> 598  -0.449859873      0
#> 599  -2.722601202      0
#> 600  -0.681077919      0
#> 601   1.293653050      0
#> 602   0.658242712      0
#> 603  -1.257009065      0
#> 604  -0.154733759      0
#> 605  -1.594848146      0
#> 606  -0.179784993      0
#> 607   2.808100892      0
#> 608   1.473323321      0
#> 609  -0.113544700      0
#> 610  -0.587786783      0
#> 611  -1.087747871      0
#> 612   0.050648801      0
#> 613  -1.195523830      0
#> 614  -0.091369283      0
#> 615  -0.033552847      0
#> 616   0.373322063      0
#> 617  -2.144132099      0
#> 618  -0.547505915      0
#> 619  -1.985005811      0
#> 620   0.052745817      0
#> 621  -0.630630395      0
#> 622  -1.003294699      0
#> 623   0.874350457      0
#> 624   1.382277293      0
#> 625   0.795199646      0
#> 626  -0.319126127      0
#> 627   0.309336934      0
#> 628  -0.992582586      0
#> 629  -0.127596641      0
#> 630   0.264147130      0
#> 631   2.204591447      0
#> 632  -0.057523147      0
#> 633   0.123893633      0
#> 634  -0.314352297      0
#> 635   0.120091338      0
#> 636   0.042246321      0
#> 637   0.704767349      0
#> 638  -0.493227804      0
#> 639   0.083701517      0
#> 640  -0.311819748      0
#> 641   0.498371891      0
#> 642   0.293250567      0
#> 643  -0.217397793      0
#> 644  -0.455494741      0
#> 645  -0.592223453      0
#> 646   0.266137724      0
#> 647  -0.126217212      0
#> 648  -0.428235749      0
#> 649  -0.768706356      0
#> 650  -0.382655614      0
#> 651  -0.980813369      0
#> 652  -1.777024940      0
#> 653   0.045314734      0
#> 654   0.772968353      0
#> 655  -0.394635861      0
#> 656  -0.975212787      0
#> 657  -0.654594987      0
#> 658   1.006970030      0
#> 659   1.108182054      0
#> 660   1.016768740      0
#> 661  -0.383606674      0
#> 662  -0.370826601      0
#> 663  -0.432927905      0
#> 664  -1.478706676      0
#> 665  -1.183288027      0
#> 666  -0.891405306      0
#> 667   0.424236214      0
#> 668  -1.352011138      0
#> 669   0.088494245      0
#> 670   0.849899756      0
#> 671   0.260400042      0
#> 672   1.006640603      0
#> 673   0.505954197      0
#> 674  -0.122651218      0
#> 675   2.095080559      0
#> 676  -0.825527617      0
#> 677   0.061080017      0
#> 678  -0.525203445      0
#> 679  -1.254564879      0
#> 680   0.073521825      0
#> 681  -0.469202229      0
#> 682  -1.136222166      0
#> 683  -2.265530521      0
#> 684   0.341732170      0
#> 685  -0.255712885      0
#> 686   1.557164248      0
#> 687  -0.591927165      0
#> 688  -0.402424459      0
#> 689  -0.710032559      0
#> 690  -0.033547761      0
#> 691   0.048235291      0
#> 692   1.031468049      0
#> 693   1.785452560      0
#> 694   0.134224590      0
#> 695  -0.453091101      0
#> 696  -0.709728830      0
#> 697  -0.496175767      0
#> 698  -0.675381306      0
#> 699   0.735155917      0
#> 700   1.003620952      0
#> 701   0.040873709      0
#> 702  -0.526881453      0
#> 703   1.565036838      0
#> 704  -0.615697917      0
#> 705  -0.153032130      0
#> 706   2.860658522      0
#> 707   0.342231497      0
#> 708   0.812572366      0
#> 709   0.278523914      0
#> 710   0.104080636      0
#> 711   1.106127289      0
#> 712   1.905899812      0
#> 713   0.390645614      0
#> 714  -0.667999184      0
#> 715   1.093569588      0
#> 716  -0.294655419      0
#> 717  -0.487655944      0
#> 718  -0.710888985      0
#> 719  -1.254863682      0
#> 720  -0.962529055      0
#> 721   1.595287823      0
#> 722  -0.433187237      0
#> 723  -1.518215944      0
#> 724   0.437624974      0
#> 725  -0.137845297      0
#> 726  -1.395645016      0
#> 727   0.036912975      0
#> 728  -0.061254031      0
#> 729  -0.377894249      0
#> 730   1.827909306      0
#> 731   0.024988296      0
#> 732   0.829528079      0
#> 733  -0.523277969      0
#> 734  -0.155734219      0
#> 735   0.190115265      0
#> 736   0.045143171      0
#> 737   0.544776158      0
#> 738  -0.109694068      0
#> 739  -0.426295832      0
#> 740   0.011520063      0
#> 741  -0.743761201      0
#> 742   0.464909197      0
#> 743  -0.182508309      0
#> 744   1.523079401      0
#> 745   0.235895244      0
#> 746  -0.881573066      0
#> 747  -0.634664832      0
#> 748   0.480375882      0
#> 749   0.640923707      0
#> 750   0.084678752      0
#> 751  -0.040670628      0
#> 752  -1.410932589      0
#> 753  -0.568125953      0
#> 754   1.241625973      0
#> 755  -0.076789544      0
#> 756   1.030062837      0
#> 757  -0.071555420      0
#> 758   0.262021355      0
#> 759   0.075152723      0
#> 760   0.813198890      0
#> 761  -0.031755082      0
#> 762   0.711186207      0
#> 763  -0.112359064      0
#> 764   0.011266946      0
#> 765  -0.900119514      0
#> 766  -1.727926427      0
#> 767   0.733629377      0
#> 768   0.151512918      0
#> 769   0.393827062      0
#> 770  -0.018951137      0
#> 771   2.197336922      0
#> 772  -0.580185936      0
#> 773  -1.245877743      0
#> 774   0.490515330      0
#> 775   0.257469555      0
#> 776   1.757889553      0
#> 777  -0.695689225      0
#> 778   1.503083367      0
#> 779   0.409939391      0
#> 780   0.253825363      0
#> 781   0.258386754      0
#> 782  -1.494967752      0
#> 783  -0.550350701      0
#> 784  -1.718615615      0
#> 785   0.932475215      0
#> 786  -0.760780812      0
#> 787  -0.130615335      0
#> 788   0.091636545      0
#> 789  -0.450932479      0
#> 790  -0.067113528      0
#> 791  -1.708195665      0
#> 792   0.964505998      0
#> 793   0.418591228      0
#> 794   1.942198629      0
#> 795  -0.277839976      0
#> 796  -0.913768226      0
#> 797  -0.710595025      0
#> 798  -1.256202968      0
#> 799   0.722475233      0
#> 800   0.273454113      0
#> 801  -0.334304580      0
#> 802   0.501656405      0
#> 803   0.884244644      0
#> 804   0.071145582      0
#> 805   1.436165819      0
#> 806  -0.962306588      0
#> 807  -0.313281581      0
#> 808  -0.389992511      0
#> 809   0.843593747      0
#> 810  -0.404635473      0
#> 811  -0.135414024      0
#> 812   1.281319868      0
#> 813   1.085200283      0
#> 814  -0.646402994      0
#> 815  -0.177764551      0
#> 816  -1.900550496      0
#> 817   0.755082105      0
#> 818   0.836783748      0
#> 819   0.491522195      0
#> 820   0.611846490      0
#> 821  -0.383166147      0
#> 822   1.666810687      0
#> 823  -0.784883096      0
#> 824   0.769415564      0
#> 825   0.991328469      0
#> 826  -2.439652531      0
#> 827   1.284510352      0
#> 828  -0.195329933      0
#> 829   0.925825836      0
#> 830  -0.060818578      0
#> 831  -0.375811791      0
#> 832  -0.795298922      0
#> 833  -0.064782460      0
#> 834  -0.752479682      0
#> 835  -0.076408340      0
#> 836   0.197718560      0
#> 837   0.459391996      0
#> 838   2.956376020      0
#> 839   0.698900278      0
#> 840  -0.137262668      0
#> 841  -1.107238413      0
#> 842   0.395173462      0
#> 843  -0.704886909      0
#> 844   0.038290599      0
#> 845   0.858068590      0
#> 846   0.365394172      0
#> 847  -1.179514011      0
#> 848  -0.135430838      0
#> 849   1.054202770      0
#> 850   0.559531201      0
#> 851  -0.666590268      0
#> 852   0.013878229      0
#> 853  -1.289557870      0
#> 854  -0.518045269      0
#> 855  -2.450192099      0
#> 856  -0.978694472      0
#> 857   0.713434352      0
#> 858  -0.383292700      0
#> 859  -0.044345029      0
#> 860  -0.059770307      0
#> 861   0.762439445      0
#> 862  -0.144856406      0
#> 863  -0.700320331      0
#> 864  -1.150104990      0
#> 865  -0.413924252      0
#> 866   1.016957703      0
#> 867  -1.994269454      0
#> 868   0.515929755      0
#> 869   0.744723875      0
#> 870  -0.972558117      0
#> 871   0.815508042      0
#> 872   0.605315267      0
#> 873  -0.210158380      0
#> 874   0.082180030      0
#> 875  -0.464970526      0
#> 876  -0.147368268      0
#> 877  -1.545031196      0
#> 878  -1.472797462      0
#> 879   1.555708434      0
#> 880   3.161524644      0
#> 881  -0.723008869      0
#> 882   0.646252402      0
#> 883   1.371166706      0
#> 884   0.277445662      0
#> 885   1.679398744      0
#> 886   1.462067163      0
#> 887  -1.320811027      0
#> 888   0.472190109      0
#> 889  -0.327837342      0
#> 890   0.301018611      0
#> 891   0.654445055      0
#> 892   0.194821145      0
#> 893   0.828965174      0
#> 894  -0.216631984      0
#> 895   0.882236572      0
#> 896   1.443747091      0
#> 897  -0.290587281      0
#> 898   0.911730408      0
#> 899  -0.220310696      0
#> 900  -0.251867430      0
#> 901  -1.121374047      0
#> 902  -1.339991211      0
#> 903   0.040757762      0
#> 904   0.482018268      0
#> 905   0.045397247      0
#> 906   1.162249470      1
#> 907  -0.797650295      1
#> 908   0.166146465      1
#> 909  -0.256423621      1
#> 910   2.050623083      1
#> 911   0.256935903      1
#> 912  -0.122998044      1
#> 913   1.517349741      1
#> 914   1.789827875      1
#> 915  -0.900389921      1
#> 916   1.590310500      1
#> 917  -0.705742768      1
#> 918   0.009868108      1
#> 919   0.251439675      1
#> 920   0.291438898      1
#> 921   1.955517865      1
#> 922   2.291224067      1
#> 923   1.587752751      1
#> 924   2.949793906      1
#> 925   1.352247140      1
#> 926   0.655387541      1
#> 927   1.096224845      1
#> 928   0.108230179      1
#> 929   0.532694601      1
#> 930   0.851396885      1
#> 931   0.415028345      1
#> 932   0.121635132      1
#> 933   2.125191654      1
#> 934   1.015462663      1
#> 935   0.732280532      1
#> 936  -0.530368687      1
#> 937   0.448197006      1
#> 938   2.051904959      1
#> 939   0.869244659      1
#> 940   1.034350074      1
#> 941  -0.122467762      1
#> 942   0.717233472      1
#> 943   1.734146350      1
#> 944  -0.209965176      1
#> 945   1.025454947      1
#> 946  -0.107510619      1
#> 947  -1.007570452      1
#> 948  -0.268433874      1
#> 949   1.182181058      1
#> 950   1.407748213      1
#> 951   0.289774175      1
#> 952   0.704430771      1
#> 953  -0.677596010      1
#> 954   0.727649441      1
#> 955   0.377641657      1
#> 956   2.170671617      1
#> 957   0.406120994      1
#> 958   0.442674451      1
#> 959  -1.288620511      1
#> 960  -0.880527784      1
#> 961   1.080418621      1
#> 962   1.855628057      1
#> 963   0.935166032      1
#> 964  -1.044808684      1
#> 965   1.565863224      1
#> 966  -0.499856652      1
#> 967   0.040260221      1
#> 968   0.206036329      1
#> 969   0.365968233      1
#> 970  -0.198650364      1
#> 971   1.374567656      1
#> 972  -0.870681875      1
#> 973   0.441810064      1
#> 974   2.092344253      1
#> 975   1.259201316      1
#> 976  -1.574734030      1
#> 977   1.898663559      1
#> 978   1.237579688      1
#> 979  -0.173952849      1
#> 980  -0.485475846      1
#> 981   0.263950591      1
#> 982   0.807969129      1
#> 983  -0.101346070      1
#> 984   1.536800412      1
#> 985  -0.329974884      1
#> 986  -0.256223340      1
#> 987   0.078445607      1
#> 988   0.053713913      1
#> 989   1.852546624      1
#> 990   1.680421592      1
#> 991  -0.770549856      1
#> 992   3.862086979      1
#> 993   1.961611694      1
#> 994   2.508784468      1
#> 995   2.201657775      1
#> 996   2.013044961      1
#> 997   0.330749083      1
#> 998   1.044920764      1
#> 999  -0.056390301      1
#> 1000  0.068634825      1