Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     1.2253152979      0
#> 2     0.0199740357      0
#> 3    -1.6029525120      0
#> 4    -0.8244743358      0
#> 5    -0.7875012989      0
#> 6    -1.3810569872      0
#> 7     0.8247883116      0
#> 8     0.9638009059      0
#> 9    -0.1263362359      0
#> 10   -0.7558532365      0
#> 11   -0.7736577675      0
#> 12    0.4274860332      0
#> 13   -0.7051102241      0
#> 14    0.9966353711      0
#> 15    0.5728597981      0
#> 16   -0.8089387117      0
#> 17   -0.3847372851      0
#> 18   -2.4343487084      0
#> 19    0.3091453427      0
#> 20    0.0611765131      0
#> 21    1.7194527747      0
#> 22    0.6528052028      0
#> 23    0.3839059098      0
#> 24    1.1230504982      0
#> 25    2.0282912633      0
#> 26   -1.2038210241      0
#> 27    0.5897383883      0
#> 28    0.2919311544      0
#> 29   -1.9086756825      0
#> 30   -0.3544754387      0
#> 31    0.8535599087      0
#> 32   -0.4636882688      0
#> 33    0.0155389638      0
#> 34   -0.3855058404      0
#> 35   -2.2256677511      0
#> 36    1.1714267540      0
#> 37   -0.3081446361      0
#> 38    1.5785187484      0
#> 39   -0.0151873149      0
#> 40   -0.1925793651      0
#> 41   -1.3024345290      0
#> 42   -0.1984444338      0
#> 43    0.5564438384      0
#> 44    0.7293504452      0
#> 45    0.2858693175      0
#> 46    0.7646863881      0
#> 47   -0.3854123213      0
#> 48   -1.0014718823      0
#> 49    0.6848400726      0
#> 50   -1.2342801312      0
#> 51   -1.0063810050      0
#> 52    0.6533971027      0
#> 53    1.1331400723      0
#> 54   -0.7300044351      0
#> 55   -1.1067209162      0
#> 56   -0.1941926531      0
#> 57   -0.3751561505      0
#> 58    0.2428396369      0
#> 59   -1.3016403991      0
#> 60    0.2180010393      0
#> 61    0.7528965802      0
#> 62   -1.3668767232      0
#> 63   -1.2073655870      0
#> 64   -0.5964570067      0
#> 65   -1.1746406452      0
#> 66    0.3674859365      0
#> 67   -1.7690251696      0
#> 68   -0.3534903094      0
#> 69    1.0219325657      0
#> 70   -1.2951786178      0
#> 71    0.2971107184      0
#> 72   -0.4085227240      0
#> 73    0.2042948021      0
#> 74   -0.2137436006      0
#> 75    0.7371418226      0
#> 76    1.2113807774      0
#> 77    0.8587844796      0
#> 78   -0.1072551680      0
#> 79   -0.9176902767      0
#> 80   -1.2692553233      0
#> 81    0.1319026481      0
#> 82    0.0089881171      0
#> 83   -0.4749828259      0
#> 84   -0.1567327259      0
#> 85   -0.7356224979      0
#> 86   -0.7046804062      0
#> 87   -1.1492980986      0
#> 88   -1.0262557560      0
#> 89    1.4331782759      0
#> 90    1.7916225014      0
#> 91   -1.4287932274      0
#> 92   -0.6688037158      0
#> 93    0.0312523749      0
#> 94   -0.9829560984      0
#> 95   -0.8772808846      0
#> 96   -2.2019440831      0
#> 97   -0.2415221401      0
#> 98   -0.2091876763      0
#> 99   -0.0789624227      0
#> 100  -0.8330806891      0
#> 101   0.6777408217      0
#> 102  -0.4865775917      0
#> 103  -1.6865651477      0
#> 104   2.6662311139      0
#> 105  -0.6610635267      0
#> 106  -1.0419133824      0
#> 107  -0.3603578806      0
#> 108   1.8051305834      0
#> 109   0.3502459706      0
#> 110   0.3093746632      0
#> 111   1.4169672235      0
#> 112   1.3520265294      0
#> 113   2.1277060781      0
#> 114   0.8679623457      0
#> 115  -0.1829111353      0
#> 116   1.0650467877      0
#> 117  -2.0313456604      0
#> 118  -0.8417923684      0
#> 119  -0.0907268774      0
#> 120   0.3089877048      0
#> 121   0.6272794180      0
#> 122  -0.5200930751      0
#> 123  -0.2248292140      0
#> 124  -0.3196030269      0
#> 125   0.4401226582      0
#> 126  -1.1101781353      0
#> 127   0.3235068851      0
#> 128   0.8800143968      0
#> 129   1.2269511421      0
#> 130   1.4468264764      0
#> 131  -0.4455148120      0
#> 132   0.5121044507      0
#> 133   0.8905644410      0
#> 134  -0.7815347615      0
#> 135  -1.0679910791      0
#> 136   1.0679711708      0
#> 137  -1.5959815929      0
#> 138  -0.7594941464      0
#> 139  -2.1612503027      0
#> 140  -0.7594998224      0
#> 141  -0.1068421259      0
#> 142  -0.1188479520      0
#> 143  -0.3879362419      0
#> 144  -1.8564484881      0
#> 145  -1.6911953309      0
#> 146   1.2096061026      0
#> 147  -1.2176552217      0
#> 148  -0.9303362463      0
#> 149   1.8390679814      0
#> 150  -0.6853866200      0
#> 151   2.3864020271      0
#> 152  -0.8935796693      0
#> 153  -0.8301416886      0
#> 154  -0.9237085181      0
#> 155   0.9252963178      0
#> 156  -1.8018492946      0
#> 157  -1.4075564068      0
#> 158  -0.6755433377      0
#> 159   0.9027020398      0
#> 160   0.5066620813      0
#> 161  -2.0922525168      0
#> 162  -0.6389222539      0
#> 163   1.4390594605      0
#> 164   1.4943742112      0
#> 165   0.0485357910      0
#> 166   0.1682268260      0
#> 167   1.7007453441      0
#> 168  -0.6523801513      0
#> 169   0.8076090936      0
#> 170   0.1476741582      0
#> 171   0.4724468053      0
#> 172  -0.1317328321      0
#> 173   0.3344212299      0
#> 174  -0.1208510137      0
#> 175  -1.7130814665      0
#> 176  -0.0907963348      0
#> 177  -0.1616300794      0
#> 178   0.4474589562      0
#> 179   1.9414184594      0
#> 180  -0.8465657204      0
#> 181   1.8173098776      0
#> 182  -1.0060950387      0
#> 183  -1.8080277758      0
#> 184   0.4470105134      0
#> 185   1.0511563998      0
#> 186   0.1462806966      0
#> 187   0.7073483339      0
#> 188   0.3671471844      0
#> 189  -0.0070394059      0
#> 190   0.1847183958      0
#> 191   1.0244139386      0
#> 192  -1.9085327995      0
#> 193   0.4066966496      0
#> 194  -1.6856737340      0
#> 195   2.2374529961      0
#> 196   0.9158156053      0
#> 197   0.4486007564      0
#> 198  -0.5233901983      0
#> 199   0.4055626318      0
#> 200  -1.0160343750      0
#> 201  -0.5757538116      0
#> 202   0.9261760643      0
#> 203  -0.3547897383      0
#> 204   1.9181552943      0
#> 205  -1.2911125517      0
#> 206   0.5026892918      0
#> 207  -0.4721794732      0
#> 208   0.7677527342      0
#> 209  -0.5039893661      0
#> 210   1.8119711570      0
#> 211   0.2675287067      0
#> 212   1.3571126717      0
#> 213   0.5393078483      0
#> 214   0.9911136531      0
#> 215   0.3705175554      0
#> 216  -0.5322722721      0
#> 217   1.1675385344      0
#> 218   1.0037064981      0
#> 219  -0.5542385846      0
#> 220   0.2566652531      0
#> 221  -0.6602302995      0
#> 222  -0.8928435287      0
#> 223  -1.1182090490      0
#> 224  -1.3134354571      0
#> 225   0.9903834664      0
#> 226   0.1862132624      0
#> 227   0.4129319079      0
#> 228  -1.0790004496      0
#> 229   0.2834072605      0
#> 230  -0.7984970206      0
#> 231   0.9631971152      0
#> 232   0.7427100262      0
#> 233   0.2243329579      0
#> 234  -1.1133374216      0
#> 235   1.4607234902      0
#> 236   0.0968514881      0
#> 237  -0.1727079741      0
#> 238   0.0259204682      0
#> 239   1.0226666686      0
#> 240  -0.2952427974      0
#> 241   0.1870996542      0
#> 242  -0.8361259601      0
#> 243   1.2505089290      0
#> 244  -0.5552447875      0
#> 245   0.4047411962      0
#> 246   1.1783549752      0
#> 247  -0.7921154395      0
#> 248  -0.5237524137      0
#> 249  -0.2257693661      0
#> 250  -1.0787462924      0
#> 251  -1.1616172054      0
#> 252   0.8379097989      0
#> 253   0.5625704753      0
#> 254   0.5070487942      0
#> 255  -0.4951711859      0
#> 256  -0.7682010837      0
#> 257   0.5272945647      0
#> 258  -1.3356074128      0
#> 259   1.1991417320      0
#> 260  -0.1425772592      0
#> 261   0.4698081649      0
#> 262  -0.3750138876      0
#> 263  -0.2361159985      0
#> 264   0.2228993434      0
#> 265  -0.4957055177      0
#> 266  -1.1532648599      0
#> 267  -0.1218462790      0
#> 268   1.6708818991      0
#> 269  -1.0210569823      0
#> 270   1.6747067971      0
#> 271  -1.0473584453      0
#> 272   0.5999660615      0
#> 273  -0.0648893746      0
#> 274  -0.7510096476      0
#> 275  -0.1135973421      0
#> 276  -0.8076707780      0
#> 277  -1.0137090676      0
#> 278  -0.9030375546      0
#> 279   2.6246538142      0
#> 280   0.9956547437      0
#> 281  -0.6978921394      0
#> 282   0.8100999270      0
#> 283  -1.1020719627      0
#> 284   2.4978069858      0
#> 285   1.6000238877      0
#> 286   0.4784737045      0
#> 287   0.1845035790      0
#> 288   0.7005448069      0
#> 289  -0.2708852830      0
#> 290  -1.2774939971      0
#> 291  -0.8973632591      0
#> 292   2.1068234835      0
#> 293  -0.2670067013      0
#> 294   0.5011867420      0
#> 295  -0.3965827945      0
#> 296   1.0251311673      0
#> 297  -1.3518350619      0
#> 298  -0.0606197831      0
#> 299  -0.5462194054      0
#> 300  -0.2210035252      0
#> 301   1.7123459518      0
#> 302   1.2834642513      0
#> 303  -1.6176756331      0
#> 304  -0.0091133530      0
#> 305   0.5969029752      0
#> 306   1.5988635656      0
#> 307  -0.2693207117      0
#> 308   1.0871693478      0
#> 309   1.2837874106      0
#> 310  -0.7360630359      0
#> 311  -2.6067326045      0
#> 312  -0.0093605879      0
#> 313   0.8450978275      0
#> 314  -1.6025364878      0
#> 315   1.3599423513      0
#> 316  -0.6177293347      0
#> 317   0.4076061284      0
#> 318  -0.6146969863      0
#> 319  -1.3319970231      0
#> 320   0.4458513773      0
#> 321   0.3320396472      0
#> 322  -1.3658588174      0
#> 323   0.0165178731      0
#> 324   0.5687117195      0
#> 325   0.6897009790      0
#> 326  -0.1531579607      0
#> 327  -1.4294122376      0
#> 328  -0.1598060284      0
#> 329  -0.6264476507      0
#> 330   0.7385663444      0
#> 331  -1.1682631845      0
#> 332   0.5838079383      0
#> 333  -1.3942512437      0
#> 334   0.0819100119      0
#> 335   1.2307713380      0
#> 336  -0.0413537378      0
#> 337   1.2262093095      0
#> 338   1.3380162104      0
#> 339  -1.2597915537      0
#> 340  -0.8321524531      0
#> 341   0.2140961275      0
#> 342   0.1865264146      0
#> 343  -0.3417661048      0
#> 344   0.2509766221      0
#> 345   0.3449219393      0
#> 346   1.2514607791      0
#> 347  -0.7454304427      0
#> 348  -0.2537749883      0
#> 349   0.7939685806      0
#> 350  -1.3665759144      0
#> 351   0.8055849398      0
#> 352   1.0350908336      0
#> 353  -1.5664932616      0
#> 354  -1.2256701501      0
#> 355   1.3464770842      0
#> 356   2.4848591974      0
#> 357  -1.1817497444      0
#> 358   0.3813658981      0
#> 359   0.6473310025      0
#> 360   0.1507430060      0
#> 361   0.0036098319      0
#> 362   0.4328751120      0
#> 363  -0.5579292121      0
#> 364  -1.8007282937      0
#> 365   1.0469613345      0
#> 366   0.6192583271      0
#> 367  -0.0778693404      0
#> 368   1.0012341317      0
#> 369  -2.4381773940      0
#> 370   0.6520185020      0
#> 371  -1.4222539145      0
#> 372  -0.9352815276      0
#> 373   0.1453920854      0
#> 374  -0.2682781315      0
#> 375   1.1182816287      0
#> 376  -0.0705935571      0
#> 377  -0.4331227599      0
#> 378  -0.6184066143      0
#> 379  -0.9991461165      0
#> 380   1.5689696865      0
#> 381  -0.9964239116      0
#> 382  -0.2778179767      0
#> 383   1.5867071638      0
#> 384  -0.5970454368      0
#> 385   0.0360332557      0
#> 386  -0.6255942365      0
#> 387  -0.0593523901      0
#> 388   0.7269712834      0
#> 389  -1.2485398150      0
#> 390  -0.6000556414      0
#> 391  -0.8814412179      0
#> 392  -0.6695405006      0
#> 393  -0.2484767215      0
#> 394   1.0594148827      0
#> 395  -1.1446123102      0
#> 396   1.1138378972      0
#> 397  -0.1318907091      0
#> 398   0.3999901458      0
#> 399  -1.3893118984      0
#> 400  -0.2788832269      0
#> 401  -0.9174456055      0
#> 402  -0.4080890039      0
#> 403  -0.1267098287      0
#> 404  -0.5707102989      0
#> 405   1.7171122846      0
#> 406  -0.5921711194      0
#> 407  -2.4101061969      0
#> 408  -0.1994924220      0
#> 409   0.2994696552      0
#> 410  -0.9628929145      0
#> 411   0.4211929141      0
#> 412   0.7494723468      0
#> 413  -0.1023580944      0
#> 414   0.7121286955      0
#> 415  -0.5462245920      0
#> 416   0.0752592318      0
#> 417  -1.0283782462      0
#> 418   0.4342932551      0
#> 419  -0.5541082026      0
#> 420  -0.8956220025      0
#> 421   0.5334154841      0
#> 422  -0.5712249438      0
#> 423  -0.6222108212      0
#> 424   1.4455610531      0
#> 425  -0.7180308294      0
#> 426  -0.6905433703      0
#> 427   0.9464039219      0
#> 428  -1.9614787568      0
#> 429  -1.0366858310      0
#> 430  -1.0447784105      0
#> 431   1.9241375946      0
#> 432   0.5207200802      0
#> 433  -1.9903566862      0
#> 434   0.3272701178      0
#> 435  -1.0918771852      0
#> 436   0.0041358027      0
#> 437   0.5854830942      0
#> 438  -1.1084558755      0
#> 439  -0.5413963936      0
#> 440  -0.5539510579      0
#> 441  -1.4374690678      0
#> 442  -0.2857203540      0
#> 443  -0.9993475513      0
#> 444  -0.0422092914      0
#> 445  -1.9660621869      0
#> 446  -0.7354159553      0
#> 447   0.9698143600      0
#> 448  -1.9947379250      0
#> 449   0.2823559780      0
#> 450  -0.4064459742      0
#> 451  -0.3410916674      0
#> 452  -0.3492469624      0
#> 453   0.1420029736      0
#> 454  -1.3199638422      0
#> 455  -0.9385920634      0
#> 456   2.4671643713      0
#> 457   2.7337993085      0
#> 458  -0.6622670342      0
#> 459  -1.3606604824      0
#> 460  -0.0144565586      0
#> 461   0.1881863780      0
#> 462  -0.6963550501      0
#> 463   1.4270390066      0
#> 464  -0.2558249478      0
#> 465   0.5373845487      0
#> 466  -0.7108430565      0
#> 467   1.6618482886      0
#> 468  -0.4118566293      0
#> 469   0.7292257676      0
#> 470  -0.2597262482      0
#> 471  -1.2761391261      0
#> 472  -0.2114849742      0
#> 473  -1.1518987684      0
#> 474  -1.4741774158      0
#> 475   0.2304236558      0
#> 476  -0.4079726426      0
#> 477  -0.2581561782      0
#> 478   0.0468060030      0
#> 479  -0.7184932876      0
#> 480  -0.8669920726      0
#> 481  -0.9490913838      0
#> 482   0.2248351674      0
#> 483   0.2821706065      0
#> 484   0.7441115699      0
#> 485   0.9916784947      0
#> 486   0.0238371489      0
#> 487   0.1184423992      0
#> 488  -3.6895796806      0
#> 489   0.1226593186      0
#> 490  -0.4600513450      0
#> 491  -1.3319799475      0
#> 492   1.6955212357      0
#> 493  -0.9518255696      0
#> 494   2.4476490950      0
#> 495  -0.6323342918      0
#> 496  -0.1504406731      0
#> 497  -0.3093576018      0
#> 498   0.0218824536      0
#> 499  -1.1868692526      0
#> 500   2.2321256400      0
#> 501  -1.8806156932      0
#> 502  -1.0195035828      0
#> 503  -2.1679347851      0
#> 504   1.2005759584      0
#> 505   0.1264327604      0
#> 506   0.2269451405      0
#> 507   0.0673662060      0
#> 508   1.2884102865      0
#> 509  -2.1626128695      0
#> 510  -1.2741695118      0
#> 511   0.1079436598      0
#> 512   2.1836677367      0
#> 513   0.5093818288      0
#> 514  -0.4669118948      0
#> 515   1.5335226454      0
#> 516   1.1201383671      0
#> 517   0.3385815521      0
#> 518  -0.5380248674      0
#> 519   0.1113668458      0
#> 520  -0.2581505734      0
#> 521  -1.6701038895      0
#> 522  -0.0005196714      0
#> 523   0.2577545359      0
#> 524  -0.0741614907      0
#> 525  -0.5884260189      0
#> 526   0.3064595700      0
#> 527  -0.1911729226      0
#> 528   0.3220866559      0
#> 529  -0.0573627923      0
#> 530  -1.8768279209      0
#> 531  -0.0399955311      0
#> 532  -0.4566892039      0
#> 533   1.2708964174      0
#> 534  -0.5488167780      0
#> 535   0.1035994322      0
#> 536  -0.4132982079      0
#> 537   0.4779101996      0
#> 538  -0.6747898616      0
#> 539   0.3342192520      0
#> 540   0.0848391780      0
#> 541   0.1696938886      0
#> 542  -0.7243656997      0
#> 543   0.5875154032      0
#> 544   2.1237168971      0
#> 545   0.4041296915      0
#> 546   0.8613771499      0
#> 547   2.7626150510      0
#> 548  -0.7817594292      0
#> 549  -0.3068244095      0
#> 550  -0.2482683113      0
#> 551  -1.6942169190      0
#> 552   0.1878654353      0
#> 553  -1.1829188564      0
#> 554   1.2675772909      0
#> 555  -0.4952128357      0
#> 556   0.1125518987      0
#> 557   1.3656756532      0
#> 558   0.7545640924      0
#> 559  -0.5244648761      0
#> 560   0.1454524759      0
#> 561   0.7427894419      0
#> 562   0.0858106465      0
#> 563  -1.3657567663      0
#> 564   0.2130059565      0
#> 565   1.3191973458      0
#> 566  -0.3801375960      0
#> 567  -1.7119510908      0
#> 568   0.5842140352      0
#> 569  -0.6994289215      0
#> 570  -0.6220353484      0
#> 571   1.0610626081      0
#> 572  -0.0246552697      0
#> 573   1.0453454240      0
#> 574  -0.8547667743      0
#> 575   0.7657289976      0
#> 576   1.0272378899      0
#> 577  -1.0663224472      0
#> 578  -1.4488539240      0
#> 579  -0.0689439449      0
#> 580  -0.4896016850      0
#> 581   1.8715811595      0
#> 582   0.3629368995      0
#> 583   0.3557064967      0
#> 584   1.3681462529      0
#> 585  -1.3240830876      0
#> 586   0.1420982607      0
#> 587  -1.2771496126      0
#> 588  -0.0947753889      0
#> 589  -1.7348856811      0
#> 590   0.0175798007      0
#> 591   0.4283635438      0
#> 592   1.5719644987      0
#> 593   1.5116797293      0
#> 594  -0.2551630194      0
#> 595   0.6641594115      0
#> 596   1.7443536466      0
#> 597   0.5064938836      0
#> 598  -0.4903190888      0
#> 599  -0.2770348204      0
#> 600  -1.3084616755      0
#> 601  -1.7812573918      0
#> 602  -0.2751904896      0
#> 603   1.6933551147      0
#> 604  -1.2539521127      0
#> 605   1.4153047005      0
#> 606   1.3157074123      0
#> 607   0.1451563164      0
#> 608  -0.3350795155      0
#> 609   0.4753344736      0
#> 610  -0.3876455676      0
#> 611   0.5351692761      0
#> 612  -1.3526744231      0
#> 613   0.2534903707      0
#> 614   0.4232372575      0
#> 615  -0.3221324145      0
#> 616   1.1252368401      0
#> 617  -1.6825292980      0
#> 618   0.1856939442      0
#> 619  -1.3887827693      0
#> 620   1.7109933826      0
#> 621  -1.0385898282      0
#> 622  -0.5738822232      0
#> 623   0.2716664565      0
#> 624  -0.9155997443      0
#> 625  -0.9607743362      0
#> 626  -0.9294639652      0
#> 627  -0.5618306885      0
#> 628  -0.7711036774      0
#> 629  -0.8374610246      0
#> 630   0.9177539934      0
#> 631   0.6022841297      0
#> 632  -0.9339535185      0
#> 633   0.3123424201      0
#> 634   0.6190584272      0
#> 635   0.1473873316      0
#> 636  -1.5841541395      0
#> 637  -0.0566944960      0
#> 638  -0.3524314474      0
#> 639   1.1384623578      0
#> 640   0.3457712360      0
#> 641   1.3754987882      0
#> 642  -1.1107798499      0
#> 643   0.0490734013      0
#> 644  -0.0937235349      0
#> 645  -0.3208641745      0
#> 646  -0.5451322685      0
#> 647  -1.4124132391      0
#> 648   0.2292707319      0
#> 649   0.4255002405      0
#> 650  -0.6214518439      0
#> 651  -2.5550723232      0
#> 652  -0.9164959012      0
#> 653  -0.0112710035      0
#> 654   0.1655780721      0
#> 655   0.0042092265      0
#> 656   0.2151586753      0
#> 657   1.9764155861      0
#> 658   2.3195908537      0
#> 659  -1.3467145607      0
#> 660   0.0662978658      0
#> 661  -1.4275686546      0
#> 662   0.9713807437      0
#> 663   0.1129312692      0
#> 664  -0.3112885934      0
#> 665  -0.5954594616      0
#> 666  -0.5295958318      0
#> 667  -0.5083106768      0
#> 668   1.7142830887      0
#> 669   1.4325718114      0
#> 670  -1.1735317932      0
#> 671   0.2179878490      0
#> 672   1.5400627571      0
#> 673   0.8766747634      0
#> 674   0.9740959056      0
#> 675   0.3362999098      0
#> 676  -0.3409798151      0
#> 677  -0.2951631785      0
#> 678  -0.2668035565      0
#> 679   0.1048480180      0
#> 680  -0.4043879311      0
#> 681   0.8653830826      0
#> 682   0.2909764791      0
#> 683   0.5038113707      0
#> 684  -0.3180150378      0
#> 685  -0.9193825606      0
#> 686   1.4634776825      0
#> 687  -0.1831083125      0
#> 688  -0.5717484355      0
#> 689  -0.4981723445      0
#> 690  -1.3978996269      0
#> 691  -0.9846698078      0
#> 692   0.9975223797      0
#> 693  -1.2622500394      0
#> 694   1.0643579679      0
#> 695   0.2883416609      0
#> 696  -2.5920463547      0
#> 697   0.5984115559      0
#> 698  -0.2656856349      0
#> 699  -0.1383041797      0
#> 700   0.0766720924      0
#> 701  -0.7296144028      0
#> 702  -0.4961777094      0
#> 703   1.3051388436      0
#> 704  -0.1810134419      0
#> 705   0.9708224837      0
#> 706  -1.3129160604      0
#> 707   0.1767384385      0
#> 708   0.4508248226      0
#> 709   1.2481561170      0
#> 710  -2.2676648675      0
#> 711  -1.7847832580      0
#> 712   1.7198516540      0
#> 713   0.4426183785      0
#> 714  -0.2248187316      0
#> 715   0.2835690501      0
#> 716   0.2609142916      0
#> 717   0.5196107132      0
#> 718   0.7018092493      0
#> 719  -0.8819714552      0
#> 720  -1.9057261734      0
#> 721   0.9744034556      0
#> 722   0.8580880189      0
#> 723   0.1149771786      0
#> 724  -0.8705421363      0
#> 725   0.5841038970      0
#> 726   0.0740400777      0
#> 727  -1.4666548021      0
#> 728   0.2544322296      0
#> 729  -0.2204393101      0
#> 730  -1.1272346037      0
#> 731  -0.5267299415      0
#> 732  -1.3754312771      0
#> 733  -3.4396742180      0
#> 734  -0.4070426627      0
#> 735   0.2823262731      0
#> 736  -0.4375917396      0
#> 737   0.7548854719      0
#> 738   1.3596649567      0
#> 739   1.6307349563      0
#> 740   0.3728773446      0
#> 741  -1.9291058703      0
#> 742   0.5940870805      0
#> 743   1.5498020169      0
#> 744  -0.2474819088      0
#> 745  -1.6873487467      0
#> 746   1.2138663459      0
#> 747  -1.4658825748      0
#> 748   1.0218316621      0
#> 749   0.8626518261      0
#> 750   0.7913622618      0
#> 751  -1.3023665409      0
#> 752  -0.2225005475      0
#> 753   0.2741427186      0
#> 754  -0.6003173637      0
#> 755  -2.1771206657      0
#> 756   1.0565111227      0
#> 757  -0.7218172508      0
#> 758  -1.9274812511      0
#> 759   0.5814403349      0
#> 760   1.6282183309      0
#> 761   1.8983159462      0
#> 762   1.5448912581      0
#> 763  -0.9979315517      0
#> 764   0.6359008006      0
#> 765   0.3313481350      0
#> 766  -0.8931554459      0
#> 767  -1.1156960198      0
#> 768  -0.2385877132      0
#> 769  -0.0150997888      0
#> 770  -1.4983071938      0
#> 771  -1.1510372679      0
#> 772   0.4310141859      0
#> 773   1.5522983009      0
#> 774  -0.3130959612      0
#> 775  -0.0979266654      0
#> 776  -2.0772998870      0
#> 777  -0.5105330472      0
#> 778  -1.2219310881      0
#> 779  -1.0540190731      0
#> 780  -1.0002449514      0
#> 781   0.2844817896      0
#> 782  -0.1317848212      0
#> 783   0.1655117250      0
#> 784   0.2737389054      0
#> 785  -0.6029824431      0
#> 786  -0.5055577459      0
#> 787   0.0751320458      0
#> 788   2.1268843666      0
#> 789   1.5863599887      0
#> 790   0.9157478870      0
#> 791   2.4276444750      0
#> 792   0.5922634541      0
#> 793  -0.3332035030      0
#> 794   0.3649254701      0
#> 795   0.3309396437      0
#> 796  -0.6474332232      0
#> 797  -1.0485356572      0
#> 798   0.3610400309      0
#> 799   0.2755964417      0
#> 800   0.2903633007      0
#> 801  -1.3858984323      0
#> 802   0.0700458832      0
#> 803   1.3010440147      0
#> 804   0.4451142101      0
#> 805  -0.4957577603      0
#> 806  -0.2560902004      0
#> 807  -1.6193918370      0
#> 808   1.6651028683      0
#> 809   1.1849312047      0
#> 810  -2.7832071119      0
#> 811  -0.0158566140      0
#> 812   0.6155118572      0
#> 813   0.8969809681      0
#> 814   2.0650096136      0
#> 815   1.0145247833      0
#> 816   0.1941495562      0
#> 817  -0.4219062998      0
#> 818  -0.4562773595      0
#> 819   2.0899690914      0
#> 820  -0.0942413482      0
#> 821   0.1746597981      0
#> 822  -0.6748109293      0
#> 823   0.1138741505      0
#> 824  -0.1609121429      0
#> 825   1.2816162479      0
#> 826  -0.7549563264      0
#> 827  -0.9963663131      0
#> 828   2.2285865672      0
#> 829   2.9591023384      0
#> 830   0.5751830995      0
#> 831  -0.4501557114      0
#> 832   0.7163732579      0
#> 833   0.4402803670      0
#> 834   0.8716546659      0
#> 835  -0.2124108639      0
#> 836   0.5779767697      0
#> 837   0.5557186660      0
#> 838   1.4890186137      0
#> 839   0.4179104451      0
#> 840  -0.0667860130      0
#> 841  -0.9590958096      0
#> 842  -1.1364178117      0
#> 843  -0.0798752116      0
#> 844   1.7291296055      0
#> 845  -0.5322232072      0
#> 846   0.1446855356      0
#> 847  -0.6054432988      0
#> 848  -0.4903550017      0
#> 849   0.1590881575      0
#> 850  -0.1690763522      0
#> 851  -1.2590054237      0
#> 852  -0.4888640895      0
#> 853   0.4396557534      0
#> 854  -0.2717929159      0
#> 855  -0.2781272879      0
#> 856  -0.8036092819      0
#> 857   0.8944378341      0
#> 858   0.4004125740      0
#> 859   0.6826821213      0
#> 860   1.2922303722      0
#> 861   1.7432333470      0
#> 862   0.3159804382      0
#> 863  -0.0501831060      0
#> 864  -0.5664325876      0
#> 865   0.1449505568      0
#> 866  -0.4911314920      0
#> 867  -0.5179463280      0
#> 868  -0.6736330896      0
#> 869  -0.0682528413      0
#> 870  -3.1256886629      0
#> 871   1.0460450880      0
#> 872   1.0428467172      0
#> 873   0.3820459924      0
#> 874  -3.1187444585      0
#> 875  -1.1841042274      0
#> 876   0.5548501877      0
#> 877   0.2820614257      0
#> 878  -0.4375786178      0
#> 879   0.8526862626      0
#> 880  -0.4441548786      0
#> 881   1.1411007267      0
#> 882  -0.9995324892      0
#> 883   1.3829399886      0
#> 884   0.2357345835      0
#> 885  -0.7902853156      0
#> 886  -0.2179009741      0
#> 887   1.3077956449      0
#> 888   0.6405061353      0
#> 889   0.0229411342      0
#> 890   0.6275694577      0
#> 891  -0.6859560797      0
#> 892   0.1147491002      0
#> 893   0.3314789491      0
#> 894  -0.5454458716      0
#> 895  -0.9991693416      0
#> 896  -1.2848751172      0
#> 897   0.6361043260      0
#> 898  -2.2443013747      0
#> 899   2.5537284659      0
#> 900   0.1013323655      0
#> 901  -0.3584936830      0
#> 902  -0.2188348111      0
#> 903   0.0938618154      0
#> 904  -1.0375379788      0
#> 905   0.8887090429      0
#> 906   1.3918108192      0
#> 907  -0.4196107571      0
#> 908  -0.4833149591      0
#> 909  -0.5830495842      0
#> 910   1.5005262519      1
#> 911   1.1195968970      1
#> 912   1.6153763053      1
#> 913  -0.1530620429      1
#> 914  -0.1758384546      1
#> 915   0.7323311319      1
#> 916   1.0856469082      1
#> 917   2.4068539698      1
#> 918   1.9308795305      1
#> 919   0.3957946724      1
#> 920  -0.9775412223      1
#> 921   1.7537278393      1
#> 922   1.0152147041      1
#> 923   0.4294055630      1
#> 924  -0.0683844970      1
#> 925  -1.3593247320      1
#> 926   0.3029992887      1
#> 927  -0.6623330351      1
#> 928   1.3097911497      1
#> 929   0.1907191010      1
#> 930   0.2526689832      1
#> 931   0.4953944276      1
#> 932   0.7605696432      1
#> 933   1.3112122310      1
#> 934   0.9631108651      1
#> 935   0.4120505284      1
#> 936   0.0214568417      1
#> 937  -0.7219712003      1
#> 938   0.0944836200      1
#> 939   0.5337201772      1
#> 940   1.4075643695      1
#> 941  -1.0403555756      1
#> 942   0.2619373035      1
#> 943  -0.5416821017      1
#> 944   2.1706356663      1
#> 945   0.2723809717      1
#> 946   1.4185277432      1
#> 947   0.6402468534      1
#> 948   1.0056136634      1
#> 949   1.9836610087      1
#> 950   2.5527313066      1
#> 951  -0.3640363369      1
#> 952   0.6400181329      1
#> 953   0.5387155946      1
#> 954   1.9198563859      1
#> 955   1.9053522596      1
#> 956   1.6909325915      1
#> 957   0.0517370019      1
#> 958   0.3578162515      1
#> 959   0.7808566676      1
#> 960  -0.1439279055      1
#> 961   0.6157216829      1
#> 962   0.7783198655      1
#> 963  -0.2104659349      1
#> 964  -0.9917530591      1
#> 965  -0.6087695506      1
#> 966  -0.2758432333      1
#> 967   0.1009914256      1
#> 968   0.6041784867      1
#> 969  -0.2294645759      1
#> 970  -0.0550189317      1
#> 971   0.5058389809      1
#> 972   0.5646428134      1
#> 973   1.8547746457      1
#> 974   1.3983494885      1
#> 975   0.7510235111      1
#> 976   1.0645878137      1
#> 977   1.2845354065      1
#> 978   1.9204790780      1
#> 979   0.8874732420      1
#> 980   3.2122290469      1
#> 981   1.4182013130      1
#> 982  -0.1639439649      1
#> 983   1.2178736104      1
#> 984  -0.0164194021      1
#> 985   0.7094184785      1
#> 986   2.1826038484      1
#> 987  -0.9201664014      1
#> 988   2.1687792659      1
#> 989  -1.7378693987      1
#> 990   1.2856157051      1
#> 991   0.9658312143      1
#> 992   0.6960467449      1
#> 993   1.7581080943      1
#> 994  -1.0679912102      1
#> 995   0.1185544521      1
#> 996   0.2184133257      1
#> 997  -0.3524351337      1
#> 998   1.7019862007      1
#> 999   0.0679261703      1
#> 1000  0.2124623877      1