Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -0.0475299947      0
#> 2    -0.0433811902      0
#> 3    -0.0969665674      0
#> 4     0.3862187504      0
#> 5     0.1278727003      0
#> 6     0.2168674157      0
#> 7    -0.0711029696      0
#> 8    -0.3882243966      0
#> 9     1.2991405969      0
#> 10    0.3356669141      0
#> 11   -1.2317693296      0
#> 12    0.6961062639      0
#> 13   -1.5898767144      0
#> 14   -1.0065961334      0
#> 15   -0.4286627772      0
#> 16    2.9895440018      0
#> 17    0.0778217977      0
#> 18    0.2839710301      0
#> 19   -0.5454205872      0
#> 20   -0.7939404534      0
#> 21   -1.6227683888      0
#> 22   -0.0820594075      0
#> 23   -0.6760586761      0
#> 24    1.2759426735      0
#> 25    0.0382034438      0
#> 26   -2.4698642925      0
#> 27    0.0698122775      0
#> 28   -1.8112904457      0
#> 29    1.0490694373      0
#> 30   -0.8592022066      0
#> 31    1.7672392961      0
#> 32    2.1483370277      0
#> 33   -1.4704359557      0
#> 34   -0.3346776213      0
#> 35   -1.9628104408      0
#> 36    0.0058952574      0
#> 37    0.1731181055      0
#> 38   -0.4052777010      0
#> 39   -0.8481628329      0
#> 40    0.1493535987      0
#> 41   -0.1884947814      0
#> 42    1.1615175536      0
#> 43   -0.7567353962      0
#> 44   -0.5301927524      0
#> 45   -0.1306714324      0
#> 46   -1.0515102841      0
#> 47    2.1382711258      0
#> 48   -0.1546441429      0
#> 49   -0.2362735636      0
#> 50    0.4782188119      0
#> 51   -0.1020380382      0
#> 52    0.4434058495      0
#> 53    0.4107156577      0
#> 54   -0.8911303303      0
#> 55    0.1268652683      0
#> 56   -0.2497456195      0
#> 57   -0.8441325741      0
#> 58    0.2939685366      0
#> 59   -1.7265186488      0
#> 60    1.5634895633      0
#> 61    1.5533743247      0
#> 62    0.3508231767      0
#> 63   -1.3481989546      0
#> 64   -0.5697669002      0
#> 65    0.4683124296      0
#> 66    0.4290674969      0
#> 67   -1.3597817893      0
#> 68   -0.1434972680      0
#> 69   -0.1741535485      0
#> 70    0.0492744544      0
#> 71   -0.3446531224      0
#> 72   -0.1309762597      0
#> 73   -0.9883498149      0
#> 74   -2.2013341763      0
#> 75   -1.1093369585      0
#> 76    0.5645490924      0
#> 77    0.6139776592      0
#> 78    0.4890566904      0
#> 79    0.4398544224      0
#> 80    2.4832803383      0
#> 81   -1.1035142670      0
#> 82    1.1359044387      0
#> 83    0.6920151581      0
#> 84    0.6507905264      0
#> 85    1.2206290517      0
#> 86   -0.4837235021      0
#> 87   -0.2008678893      0
#> 88   -0.2934748682      0
#> 89    0.3683599900      0
#> 90   -1.1753771019      0
#> 91    1.7867747818      0
#> 92   -0.3453019500      0
#> 93    0.7314991292      0
#> 94   -0.8740219992      0
#> 95   -1.0351865780      0
#> 96    0.9299645214      0
#> 97    0.3799014661      0
#> 98   -1.0184528336      0
#> 99   -0.3390882925      0
#> 100   0.0009153082      0
#> 101   1.1493466394      0
#> 102  -1.7855208898      0
#> 103   0.5535461083      0
#> 104  -0.5056737884      0
#> 105   0.1793694485      0
#> 106   1.0089930367      0
#> 107  -0.2335673829      0
#> 108  -1.5426054564      0
#> 109   0.6885600217      0
#> 110  -0.8180456151      0
#> 111  -0.6574905893      0
#> 112  -1.6909010684      0
#> 113   0.9325382777      0
#> 114   0.4794224884      0
#> 115  -1.0082452776      0
#> 116  -1.0641753799      0
#> 117   0.3449540532      0
#> 118  -0.1857140046      0
#> 119   2.3006665934      0
#> 120   0.3879726881      0
#> 121  -0.7855308751      0
#> 122   0.1752686995      0
#> 123   1.4852483833      0
#> 124  -0.7763429721      0
#> 125   1.6159533746      0
#> 126  -0.8474747797      0
#> 127   0.8690231541      0
#> 128   0.6345869625      0
#> 129   0.3102555536      0
#> 130   0.5094469278      0
#> 131  -0.5330593663      0
#> 132   0.0322633405      0
#> 133   0.8272995396      0
#> 134   0.1798448298      0
#> 135   0.1969299529      0
#> 136   0.5011096643      0
#> 137  -0.3727947495      0
#> 138   0.7595967832      0
#> 139   0.6562494014      0
#> 140  -0.1281893621      0
#> 141  -0.9655823882      0
#> 142   1.2377409516      0
#> 143   0.5989491273      0
#> 144  -0.7889665983      0
#> 145   0.5666971566      0
#> 146   0.5178106623      0
#> 147  -0.3111555895      0
#> 148   1.3553442926      0
#> 149   0.2130870944      0
#> 150   1.1389847876      0
#> 151   0.2110501236      0
#> 152   0.1702406852      0
#> 153  -0.7023673750      0
#> 154  -0.8814823050      0
#> 155  -2.0601440249      0
#> 156   1.1158324628      0
#> 157  -0.5897230108      0
#> 158  -0.4790476607      0
#> 159  -1.0658766442      0
#> 160  -0.6318969722      0
#> 161   0.3928454708      0
#> 162   0.2707383676      0
#> 163  -0.9206881269      0
#> 164   0.3017919899      0
#> 165  -0.2199172388      0
#> 166  -1.0048871826      0
#> 167   0.2638634891      0
#> 168   0.1132413782      0
#> 169   0.8000616600      0
#> 170  -1.3833942260      0
#> 171  -0.1081404611      0
#> 172   1.1795624796      0
#> 173  -0.6318197612      0
#> 174   2.5271582756      0
#> 175   0.1547632799      0
#> 176   0.4602216437      0
#> 177  -0.3887703314      0
#> 178  -0.3403478263      0
#> 179  -0.9967790922      0
#> 180  -1.1326504293      0
#> 181  -0.7828012188      0
#> 182  -0.8289624299      0
#> 183   0.8172127694      0
#> 184   1.1766384690      0
#> 185  -0.4861539676      0
#> 186  -1.4628210925      0
#> 187   0.3919296745      0
#> 188   0.4126237196      0
#> 189  -0.5284161245      0
#> 190   2.0670338824      0
#> 191  -0.3122447804      0
#> 192  -2.2171995721      0
#> 193   0.6071976281      0
#> 194  -0.9702250257      0
#> 195   0.9533876903      0
#> 196  -0.5128388360      0
#> 197   1.0823482451      0
#> 198  -2.1374673334      0
#> 199  -1.1022059148      0
#> 200  -0.4198962939      0
#> 201  -0.9960533994      0
#> 202  -0.9579474842      0
#> 203  -0.3817457961      0
#> 204  -0.3274550933      0
#> 205  -0.9259406651      0
#> 206  -0.8975391397      0
#> 207  -0.8997238337      0
#> 208   0.1003400904      0
#> 209  -1.2061025896      0
#> 210   0.3204884872      0
#> 211   0.4901358284      0
#> 212   0.3705198742      0
#> 213   0.3155759069      0
#> 214  -1.3748315555      0
#> 215   1.8849825633      0
#> 216  -1.0155828298      0
#> 217  -0.6346690917      0
#> 218  -1.8864606417      0
#> 219   0.6281144431      0
#> 220   1.1675516196      0
#> 221   0.7108750006      0
#> 222  -0.0478418999      0
#> 223   0.7067194230      0
#> 224   0.1425186461      0
#> 225   0.3592266316      0
#> 226  -0.2172380878      0
#> 227   0.1419077384      0
#> 228  -0.3424033796      0
#> 229  -0.5633474113      0
#> 230   1.1966217401      0
#> 231   0.3884089194      0
#> 232   0.9760697675      0
#> 233   1.2583974583      0
#> 234   0.9862775789      0
#> 235   1.5238723983      0
#> 236  -0.3758567961      0
#> 237  -0.6822302650      0
#> 238   0.4471378236      0
#> 239   1.2537745914      0
#> 240  -1.7021124276      0
#> 241   0.4927352854      0
#> 242  -0.1267576005      0
#> 243  -0.5889238883      0
#> 244   0.6162964661      0
#> 245  -0.7552923138      0
#> 246   0.4335774654      0
#> 247  -0.5974038238      0
#> 248  -0.4952873261      0
#> 249  -0.2052933237      0
#> 250   0.2562846614      0
#> 251   1.3118404234      0
#> 252   0.4615157633      0
#> 253   0.8380618304      0
#> 254  -1.3502833524      0
#> 255   0.0994724292      0
#> 256   1.4188494362      0
#> 257  -0.6269052297      0
#> 258   0.7244170962      0
#> 259   0.8973732584      0
#> 260  -0.0792732413      0
#> 261  -0.6187439687      0
#> 262  -2.1442766456      0
#> 263   0.3296932518      0
#> 264   0.1545209764      0
#> 265  -0.8146532385      0
#> 266   1.0106735850      0
#> 267   1.3387732463      0
#> 268   1.6476750680      0
#> 269   0.0591050356      0
#> 270  -0.0094148540      0
#> 271   1.1146292670      0
#> 272  -0.3995644123      0
#> 273   1.5250437940      0
#> 274   0.2108201470      0
#> 275   0.3681763676      0
#> 276  -0.1402296432      0
#> 277  -0.6727329636      0
#> 278   0.8079069261      0
#> 279   0.1155845241      0
#> 280   0.8071714522      0
#> 281  -0.0894056570      0
#> 282  -0.4847389384      0
#> 283   1.6581444024      0
#> 284  -0.4614336067      0
#> 285   0.1426770927      0
#> 286  -0.2650816765      0
#> 287   0.5296345063      0
#> 288  -0.1046931817      0
#> 289  -0.5945802242      0
#> 290  -0.4174815220      0
#> 291   0.1434406867      0
#> 292   0.5976716991      0
#> 293   0.1863045623      0
#> 294  -0.4451148777      0
#> 295   0.2555665780      0
#> 296   0.3456141374      0
#> 297  -0.2319710048      0
#> 298  -0.2095557674      0
#> 299   1.2872324095      0
#> 300  -0.0033895667      0
#> 301  -0.5552031402      0
#> 302   0.5822926089      0
#> 303  -0.5997419073      0
#> 304  -0.6143261033      0
#> 305  -2.5096131522      0
#> 306  -0.2432799038      0
#> 307  -0.0732050414      0
#> 308  -2.4416565197      0
#> 309   0.9062714799      0
#> 310  -1.0945590910      0
#> 311  -0.7975879244      0
#> 312   0.7178972073      0
#> 313  -0.2538990524      0
#> 314   0.4356155057      0
#> 315  -1.6585692427      0
#> 316  -0.4352967646      0
#> 317  -1.2345658396      0
#> 318  -2.5196161210      0
#> 319   0.3593595781      0
#> 320  -1.5156773190      0
#> 321  -0.1883369700      0
#> 322   0.7456676502      0
#> 323  -0.3069248031      0
#> 324   0.5932772761      0
#> 325   0.1337440670      0
#> 326  -0.1790325707      0
#> 327  -0.1962901751      0
#> 328   0.2373829630      0
#> 329  -0.8978986652      0
#> 330  -1.4288415718      0
#> 331   1.7218538884      0
#> 332  -0.4357002516      0
#> 333  -0.0158124697      0
#> 334  -1.4165489799      0
#> 335   0.6884999749      0
#> 336  -0.1756304822      0
#> 337   2.3087587417      0
#> 338  -0.5746907450      0
#> 339  -0.4218376423      0
#> 340  -0.2588026288      0
#> 341   0.0178601583      0
#> 342   0.5234840859      0
#> 343   0.2541383522      0
#> 344  -0.3799923224      0
#> 345  -1.3688044036      0
#> 346   0.8094111216      0
#> 347  -1.3506376520      0
#> 348  -0.2056823666      0
#> 349   0.7580932385      0
#> 350  -0.5360931971      0
#> 351   0.6863094395      0
#> 352  -0.1096090121      0
#> 353  -1.1982348824      0
#> 354  -0.5124621825      0
#> 355   0.1756635359      0
#> 356  -0.7043325440      0
#> 357   0.9687596150      0
#> 358   0.2940662070      0
#> 359   0.5128540922      0
#> 360   1.1882052001      0
#> 361  -0.3327147752      0
#> 362   0.6318651326      0
#> 363  -0.5895141793      0
#> 364  -3.8757250111      0
#> 365   0.7853559156      0
#> 366  -0.0657229283      0
#> 367  -0.3765640317      0
#> 368   0.4888652042      0
#> 369  -0.1748905992      0
#> 370  -0.8259467684      0
#> 371   0.5091012448      0
#> 372  -1.2130048849      0
#> 373  -0.5076175857      0
#> 374  -0.6073910157      0
#> 375  -0.2033534733      0
#> 376   0.2153033116      0
#> 377   1.3305912147      0
#> 378  -0.1647883894      0
#> 379  -1.7990299126      0
#> 380  -0.5002666734      0
#> 381   0.0831743849      0
#> 382   0.1807591244      0
#> 383  -1.3217100359      0
#> 384  -0.7080820210      0
#> 385  -1.9085366694      0
#> 386   1.0175345981      0
#> 387  -0.3276498152      0
#> 388   1.1335627531      0
#> 389   0.3613074150      0
#> 390  -2.0338493276      0
#> 391   0.4272473005      0
#> 392  -0.0937740679      0
#> 393   0.2766823537      0
#> 394  -0.3380197099      0
#> 395  -0.5981590957      0
#> 396  -0.9876133373      0
#> 397  -1.1542240978      0
#> 398   1.7331810128      0
#> 399   0.6053819954      0
#> 400  -0.4600574207      0
#> 401  -1.6524073088      0
#> 402   1.4203194861      0
#> 403  -0.3426242149      0
#> 404  -0.3741846224      0
#> 405  -1.3859495546      0
#> 406  -2.1111880008      0
#> 407  -1.6442293877      0
#> 408   0.9109716045      0
#> 409   0.1734114472      0
#> 410   1.1225522989      0
#> 411  -1.1696240543      0
#> 412  -1.4490218348      0
#> 413   2.0137622932      0
#> 414  -0.6104764382      0
#> 415  -0.9129323416      0
#> 416  -0.5877542413      0
#> 417  -0.2888685727      0
#> 418   0.8462450332      0
#> 419   2.6372780967      0
#> 420  -0.4181653592      0
#> 421   1.6735027022      0
#> 422  -1.3542003140      0
#> 423  -2.1547893718      0
#> 424  -0.9739997462      0
#> 425  -0.9691922164      0
#> 426   0.3502601721      0
#> 427  -0.0407976201      0
#> 428  -2.5363299522      0
#> 429  -0.3573388700      0
#> 430  -0.4776937148      0
#> 431   1.1250443167      0
#> 432  -0.4630423051      0
#> 433   1.1328786299      0
#> 434   1.3783743535      0
#> 435   0.7707831539      0
#> 436   0.8391478489      0
#> 437   0.3768084753      0
#> 438   1.5741833034      0
#> 439   0.3449372319      0
#> 440  -0.6827442522      0
#> 441   1.4062198233      0
#> 442  -0.6241322620      0
#> 443  -1.2941227701      0
#> 444   0.5890587494      0
#> 445  -0.7890556317      0
#> 446  -0.2303881879      0
#> 447   1.0185407708      0
#> 448   0.6367496902      0
#> 449   0.1310564191      0
#> 450  -2.3306219387      0
#> 451   0.0422404034      0
#> 452  -0.8962198776      0
#> 453   1.6637587262      0
#> 454  -2.8812664632      0
#> 455  -0.4456319187      0
#> 456  -0.3233687699      0
#> 457   0.0443527039      0
#> 458  -1.2816042672      0
#> 459  -0.9716396519      0
#> 460   0.4233301803      0
#> 461  -0.4468860809      0
#> 462  -0.1006851003      0
#> 463  -1.1779949831      0
#> 464  -0.9002970649      0
#> 465  -0.4661850930      0
#> 466  -0.4909868962      0
#> 467   0.2952980829      0
#> 468   1.2577947087      0
#> 469  -0.9385231206      0
#> 470   0.4080549189      0
#> 471   0.4454730519      0
#> 472  -0.7251555108      0
#> 473  -0.6477790904      0
#> 474  -0.1818045809      0
#> 475  -0.8574224678      0
#> 476   1.9383061706      0
#> 477   0.7376542888      0
#> 478  -1.1997545116      0
#> 479   0.3448954329      0
#> 480  -0.0568763479      0
#> 481  -0.9501442727      0
#> 482   0.5750534987      0
#> 483  -0.0009402850      0
#> 484   0.3783872286      0
#> 485  -1.6426406155      0
#> 486  -0.6826316135      0
#> 487   0.2386627663      0
#> 488   0.1489026932      0
#> 489   1.7146268223      0
#> 490  -1.3084627518      0
#> 491   0.5033469072      0
#> 492   0.2507128752      0
#> 493   0.3002813780      0
#> 494   0.9204346193      0
#> 495  -0.4733179208      0
#> 496   0.3805678677      0
#> 497  -1.2399228151      0
#> 498   0.4162794990      0
#> 499   1.2327078723      0
#> 500  -0.0928243436      0
#> 501   0.9500067630      0
#> 502  -1.5087373031      0
#> 503  -0.1867631565      0
#> 504  -1.1646354199      0
#> 505  -2.0090309669      0
#> 506  -2.3775292487      0
#> 507  -0.6545270398      0
#> 508   0.1841308219      0
#> 509   1.0879592649      0
#> 510   0.2544926210      0
#> 511   0.9258596140      0
#> 512  -1.7699675775      0
#> 513  -0.2313448789      0
#> 514   0.6389645031      0
#> 515   1.4837276805      0
#> 516   0.7974368097      0
#> 517   0.2971154857      0
#> 518   0.8023723677      0
#> 519   0.4949941600      0
#> 520   1.5875268969      0
#> 521   0.6495509572      0
#> 522   0.2684944014      0
#> 523  -0.5484348846      0
#> 524  -1.6731781201      0
#> 525  -0.4789732975      0
#> 526  -0.5583395104      0
#> 527  -0.2035010779      0
#> 528   0.4089365563      0
#> 529  -0.9332829421      0
#> 530   1.3875310543      0
#> 531  -0.0499120673      0
#> 532   0.1209990844      0
#> 533   0.0022518181      0
#> 534  -1.8992877430      0
#> 535   0.6177813312      0
#> 536  -0.7249017523      0
#> 537   0.2552447918      0
#> 538   0.6326117781      0
#> 539   1.2085828365      0
#> 540  -0.2641091637      0
#> 541  -1.5917202382      0
#> 542  -0.4511863509      0
#> 543  -0.1548337592      0
#> 544   1.0705579002      0
#> 545   1.4861941279      0
#> 546  -0.9265097942      0
#> 547  -0.8488107059      0
#> 548   0.8054351691      0
#> 549   0.5393620537      0
#> 550   0.9847097947      0
#> 551  -0.2070305138      0
#> 552   0.4435721566      0
#> 553  -0.8598012123      0
#> 554   0.5709527238      0
#> 555   0.7777864425      0
#> 556  -0.4380692144      0
#> 557  -0.0686257906      0
#> 558  -0.8793916797      0
#> 559   1.0001698293      0
#> 560  -1.0072542158      0
#> 561   0.7975152184      0
#> 562  -1.5511029761      0
#> 563  -0.2592322174      0
#> 564   0.7982448935      0
#> 565  -1.0390696901      0
#> 566  -0.4105687244      0
#> 567   0.1650847703      0
#> 568  -0.4214151446      0
#> 569  -0.1745759499      0
#> 570  -0.6258185979      0
#> 571  -0.3562854424      0
#> 572  -0.0957415206      0
#> 573   0.6099772765      0
#> 574  -0.4970649494      0
#> 575  -0.7165811963      0
#> 576   0.5130463291      0
#> 577   1.0005806075      0
#> 578   0.9725627404      0
#> 579   0.9256273055      0
#> 580  -0.4411399052      0
#> 581  -0.5792731009      0
#> 582  -0.9586635398      0
#> 583   0.4082051677      0
#> 584   1.1201138625      0
#> 585  -1.2058420902      0
#> 586   0.1402324805      0
#> 587   0.0411109176      0
#> 588  -0.2591198388      0
#> 589  -1.4488485184      0
#> 590  -0.0503474972      0
#> 591  -0.5501170943      0
#> 592   2.5068640253      0
#> 593  -1.7400816915      0
#> 594  -0.6368459657      0
#> 595   0.6817406846      0
#> 596   1.1868448601      0
#> 597  -0.5544836391      0
#> 598   0.1621244745      0
#> 599   1.3132567902      0
#> 600   1.1173792180      0
#> 601  -0.1942092295      0
#> 602   2.3808455441      0
#> 603  -0.8229797110      0
#> 604   0.0419601940      0
#> 605  -0.5115693761      0
#> 606   0.4433945514      0
#> 607  -1.0539467340      0
#> 608   0.3541025178      0
#> 609   1.2444800611      0
#> 610   1.3377185702      0
#> 611   0.9699141026      0
#> 612  -0.0227708375      0
#> 613  -1.0370486204      0
#> 614   0.5138339792      0
#> 615  -1.3052415029      0
#> 616   0.1434794495      0
#> 617  -0.7373804843      0
#> 618   0.1614596476      0
#> 619  -0.1613789028      0
#> 620   1.2006508133      0
#> 621  -0.9493376367      0
#> 622  -1.9695711337      0
#> 623   0.8561491806      0
#> 624   0.2051823519      0
#> 625  -0.0689197067      0
#> 626   1.1178158978      0
#> 627   0.8116162648      0
#> 628   1.3874661757      0
#> 629  -1.8566156707      0
#> 630   0.1850259609      0
#> 631  -0.2602349169      0
#> 632   2.4576858419      0
#> 633   0.5141919614      0
#> 634  -0.1602359406      0
#> 635   0.4492900620      0
#> 636  -2.4669874610      0
#> 637   0.0878877402      0
#> 638   0.5777002971      0
#> 639  -0.2360298037      0
#> 640   1.1680884343      0
#> 641  -0.0057520372      0
#> 642  -0.9686064463      0
#> 643   0.0448757075      0
#> 644   0.6257018528      0
#> 645  -0.0479382983      0
#> 646   0.1077590204      0
#> 647  -1.1029463550      0
#> 648   1.2420642247      0
#> 649   1.5610921306      0
#> 650  -0.5789136518      0
#> 651  -0.1835774652      0
#> 652   0.1105346157      0
#> 653  -0.0968898629      0
#> 654   0.6628190869      0
#> 655  -0.7181076783      0
#> 656   1.0753636930      0
#> 657   0.9907962080      0
#> 658  -0.7784436072      0
#> 659  -0.0390111927      0
#> 660  -0.0605208081      0
#> 661   0.1223068228      0
#> 662   0.3602034344      0
#> 663   0.2507968583      0
#> 664  -0.4888979433      0
#> 665   0.3041110844      0
#> 666  -0.3845780097      0
#> 667   0.1576393218      0
#> 668   1.4494604536      0
#> 669   2.1102149727      0
#> 670  -0.8637634256      0
#> 671   0.3797884059      0
#> 672   0.8102034317      0
#> 673   1.3066730661      0
#> 674  -0.9945581869      0
#> 675  -0.6038400204      0
#> 676   0.8244298840      0
#> 677  -1.8214062292      0
#> 678  -0.5796556440      0
#> 679   1.6866419115      0
#> 680   0.4244348172      0
#> 681   0.4123291708      0
#> 682  -0.7287771423      0
#> 683  -0.1887794200      0
#> 684  -0.4833866428      0
#> 685  -0.2883378896      0
#> 686  -0.2972236265      0
#> 687   0.0039637569      0
#> 688   1.4551377488      0
#> 689   1.0450301863      0
#> 690   0.8741006867      0
#> 691   0.5689341863      0
#> 692   1.2294785011      0
#> 693  -0.0071781940      0
#> 694   0.0494891426      0
#> 695  -0.9702583578      0
#> 696   0.2133828568      0
#> 697   0.1978932552      0
#> 698  -0.4463616025      0
#> 699   0.2921179312      0
#> 700   0.4986334523      0
#> 701  -0.3882966388      0
#> 702  -1.7694816728      0
#> 703   0.4379408487      0
#> 704  -0.2015278785      0
#> 705  -1.4797289375      0
#> 706  -0.5335129470      0
#> 707  -0.1452803607      0
#> 708  -0.6360994948      0
#> 709  -0.2596054499      0
#> 710  -1.5967287804      0
#> 711   0.3763428661      0
#> 712  -0.0254448258      0
#> 713   0.9864495947      0
#> 714   0.4368163213      0
#> 715   0.4264306070      0
#> 716  -0.5314811271      0
#> 717  -0.3776711824      0
#> 718   0.3393137564      0
#> 719   0.6438219234      0
#> 720   0.1566561507      0
#> 721   0.3207450702      0
#> 722   1.2461225626      0
#> 723   1.4228058427      0
#> 724  -1.2799419939      0
#> 725  -0.2546921480      0
#> 726  -1.1527209784      0
#> 727  -0.0978922449      0
#> 728  -1.5528289733      0
#> 729  -1.0106700072      0
#> 730  -0.1789656323      0
#> 731   0.3726250436      0
#> 732   0.1644221604      0
#> 733   1.3817826790      0
#> 734   0.9644833137      0
#> 735  -1.1812170121      0
#> 736   0.7475850925      0
#> 737  -1.8483158413      0
#> 738  -1.3585912003      0
#> 739  -0.5301849796      0
#> 740  -0.9577527355      0
#> 741  -0.1183753855      0
#> 742   0.5039796506      0
#> 743   0.0400300036      0
#> 744  -0.2061377322      0
#> 745  -0.5236240627      0
#> 746  -0.8337071919      0
#> 747  -0.4676665962      0
#> 748  -1.2968695950      0
#> 749  -0.1462087532      0
#> 750  -0.4752744959      0
#> 751   0.3558512076      0
#> 752   0.0295605674      0
#> 753  -1.4728561387      0
#> 754  -0.0344842393      0
#> 755  -0.4756321790      0
#> 756  -1.0560867816      0
#> 757  -0.4325835963      0
#> 758   0.3311730662      0
#> 759  -0.2638874450      0
#> 760   0.6043676312      0
#> 761   0.2490558952      0
#> 762   0.1235554870      0
#> 763   0.0956292396      0
#> 764   1.2710111586      0
#> 765   0.9648801681      0
#> 766   0.6244508270      0
#> 767  -1.8455214607      0
#> 768   0.0356688155      0
#> 769  -0.4319851150      0
#> 770   0.9942451040      0
#> 771  -1.4628930654      0
#> 772  -0.8716965422      0
#> 773   0.3906252707      0
#> 774   0.2736128283      0
#> 775  -0.0498267903      0
#> 776  -0.1006569208      0
#> 777   2.0284287293      0
#> 778   2.0160068121      0
#> 779   1.3317398368      0
#> 780  -1.0364886588      0
#> 781  -1.4792588597      0
#> 782   1.5905405535      0
#> 783  -2.0233416634      0
#> 784  -1.8336420247      0
#> 785  -1.0302785179      0
#> 786   2.4989440503      0
#> 787  -0.8222633257      0
#> 788  -0.6769298396      0
#> 789   0.2356515947      0
#> 790   0.4982933643      0
#> 791   0.1467769254      0
#> 792  -0.2157278182      0
#> 793  -0.9344782968      0
#> 794   0.1393454591      0
#> 795   1.0679265635      0
#> 796  -2.1077478775      0
#> 797   1.3694648206      0
#> 798  -0.5352929767      0
#> 799  -0.7941979457      0
#> 800   0.2330238013      0
#> 801  -0.4214494325      0
#> 802   0.0656508904      0
#> 803   0.5400346372      0
#> 804   0.0935445004      0
#> 805  -0.7133552479      0
#> 806   0.3126848989      0
#> 807   1.1823261789      0
#> 808   1.0714981846      0
#> 809   0.0677793615      0
#> 810  -0.2428831078      0
#> 811   0.7030412355      0
#> 812   1.1757739998      0
#> 813   0.0269345497      0
#> 814  -0.2547689062      0
#> 815   0.1540797549      0
#> 816   1.0179986128      0
#> 817  -0.5149310565      0
#> 818   0.3210258450      0
#> 819   0.0790275368      0
#> 820  -1.8103243607      0
#> 821  -1.8094125107      0
#> 822  -2.2976449434      0
#> 823  -0.6965999764      0
#> 824  -0.5240940907      0
#> 825   1.4008272652      0
#> 826  -0.0381416763      0
#> 827  -0.3187059874      0
#> 828   0.1960133493      0
#> 829  -0.9819062643      0
#> 830  -0.4196050240      0
#> 831   0.3932962659      0
#> 832   0.0962739991      0
#> 833   1.4289746548      0
#> 834  -0.7141824529      0
#> 835  -2.8293114658      0
#> 836  -0.0838097315      0
#> 837  -0.6195409654      0
#> 838  -0.7439429508      0
#> 839  -0.3867923585      0
#> 840  -1.5443279591      0
#> 841  -0.5000641269      0
#> 842   0.0429338081      0
#> 843   0.3466246666      0
#> 844   0.7030990848      0
#> 845   1.4485762050      0
#> 846   0.1432857195      0
#> 847  -1.1710304712      0
#> 848   0.4161115732      0
#> 849   2.4423791656      0
#> 850   0.1309456188      0
#> 851   1.7286503060      0
#> 852   1.5290271008      0
#> 853  -2.2812605376      0
#> 854  -1.9307903488      0
#> 855   0.3932767442      0
#> 856   0.0378799538      0
#> 857   0.5961236716      0
#> 858   2.7860670467      0
#> 859   0.0247833818      0
#> 860  -1.4966102257      0
#> 861   1.0462156196      0
#> 862  -0.2632981367      0
#> 863   0.3059381964      0
#> 864  -0.8729845718      0
#> 865   1.5781000654      0
#> 866  -0.5720420100      0
#> 867  -2.6328319498      0
#> 868   0.1899765749      0
#> 869  -0.4040963320      0
#> 870   1.4020877819      0
#> 871   0.6871435547      0
#> 872  -1.5535240830      0
#> 873   1.5522089119      0
#> 874   0.1152730490      0
#> 875   0.4002073093      0
#> 876   0.5658249026      0
#> 877   0.4837893057      0
#> 878   0.5466088634      0
#> 879   0.0526927059      0
#> 880   0.8022249872      0
#> 881   0.8502589317      0
#> 882   0.2985914979      0
#> 883   0.8474552611      0
#> 884   0.9644360660      0
#> 885  -0.0068819271      0
#> 886   0.1544829741      0
#> 887  -0.5571645711      1
#> 888   0.3784499657      1
#> 889   1.0426882121      1
#> 890   0.2712871805      1
#> 891   1.7187598627      1
#> 892  -0.2748137678      1
#> 893  -1.0274667102      1
#> 894   2.3289829100      1
#> 895   2.5247896386      1
#> 896   0.7214586668      1
#> 897   1.6576818729      1
#> 898   0.0880884476      1
#> 899  -0.7756744912      1
#> 900   0.4738147767      1
#> 901  -2.3331715973      1
#> 902   0.9738255322      1
#> 903   0.6900516377      1
#> 904   1.7645051304      1
#> 905  -0.5899754967      1
#> 906   0.5414970733      1
#> 907   1.7544163633      1
#> 908  -0.1593124586      1
#> 909   1.4785832372      1
#> 910   1.5996127859      1
#> 911   1.8863129877      1
#> 912   0.5444840619      1
#> 913  -0.8112396901      1
#> 914  -0.3183217525      1
#> 915  -0.1039329212      1
#> 916   0.8251300067      1
#> 917   0.8464983131      1
#> 918   1.4444321582      1
#> 919   1.1106215767      1
#> 920  -0.8518486129      1
#> 921   0.3930864834      1
#> 922   0.5419372526      1
#> 923   1.1943083509      1
#> 924   0.8873771143      1
#> 925   2.1569189637      1
#> 926   1.7996438843      1
#> 927   0.2893724265      1
#> 928   0.3114267704      1
#> 929   0.5602031244      1
#> 930   1.2054582203      1
#> 931   0.8210897703      1
#> 932   1.8335838333      1
#> 933   0.2260785195      1
#> 934   1.7632713071      1
#> 935   1.1578422137      1
#> 936   1.3572474815      1
#> 937   0.5342910555      1
#> 938   0.8451369679      1
#> 939  -0.1371650044      1
#> 940   1.1287509642      1
#> 941  -0.6028622816      1
#> 942   1.0947715437      1
#> 943   0.5632516817      1
#> 944  -0.8113006127      1
#> 945   0.8351738888      1
#> 946   1.3137556061      1
#> 947   1.2414319049      1
#> 948   0.1172215941      1
#> 949  -0.4831984973      1
#> 950   1.6528242346      1
#> 951   1.4449222652      1
#> 952   0.9809894079      1
#> 953  -0.2460527417      1
#> 954   1.8524604104      1
#> 955   1.3216520090      1
#> 956   0.8695980747      1
#> 957   0.8339153880      1
#> 958   0.1722889367      1
#> 959   1.1200013843      1
#> 960  -0.1360487716      1
#> 961  -0.1173803852      1
#> 962   2.0953575519      1
#> 963   2.1169269520      1
#> 964  -0.4604434763      1
#> 965   1.7029256894      1
#> 966   0.8701069575      1
#> 967  -0.8324726143      1
#> 968  -0.0556029192      1
#> 969   0.6453343050      1
#> 970   1.2233407879      1
#> 971   0.9846888434      1
#> 972   1.2160198372      1
#> 973   0.3380031085      1
#> 974  -0.2783506721      1
#> 975   2.0421504697      1
#> 976   0.3237913584      1
#> 977   1.0734785081      1
#> 978   1.0474290661      1
#> 979  -1.6597709824      1
#> 980   1.2691628661      1
#> 981   2.2586933736      1
#> 982  -0.9367778885      1
#> 983   0.4803707349      1
#> 984   0.8328991657      1
#> 985   2.1965708291      1
#> 986   1.4351756363      1
#> 987   0.4985061763      1
#> 988   0.6538015205      1
#> 989  -0.5806968880      1
#> 990   0.1784767827      1
#> 991   0.3547838833      1
#> 992  -0.6801418750      1
#> 993   0.9080425945      1
#> 994   0.4108414698      1
#> 995   0.4646497322      1
#> 996   1.9545559415      1
#> 997   1.0339377971      1
#> 998  -0.2239913673      1
#> 999   0.0393604143      1
#> 1000  0.9066977024      1