Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.837152448      0
#> 2    -0.884114560      0
#> 3     1.954050437      0
#> 4     1.544103761      0
#> 5     1.298361106      0
#> 6    -0.978394039      0
#> 7    -0.196543042      0
#> 8    -0.311683646      0
#> 9     0.310293174      0
#> 10    0.216436251      0
#> 11    0.480084368      0
#> 12   -1.042895584      0
#> 13    0.189347687      0
#> 14   -0.556976230      0
#> 15    0.021011160      0
#> 16   -1.126078355      0
#> 17   -0.905100928      0
#> 18   -0.720489195      0
#> 19    1.803667952      0
#> 20    0.780140944      0
#> 21   -1.573940790      0
#> 22   -1.987498465      0
#> 23    0.662720313      0
#> 24    0.267401983      0
#> 25    0.128290291      0
#> 26   -1.156623810      0
#> 27   -0.730964523      0
#> 28   -0.755515755      0
#> 29    0.480576735      0
#> 30    0.694945227      0
#> 31    1.095928141      0
#> 32   -0.016426999      0
#> 33    0.351934080      0
#> 34   -0.665822689      0
#> 35   -0.605391566      0
#> 36   -2.364051729      0
#> 37   -1.683568015      0
#> 38   -1.033323774      0
#> 39    0.956780793      0
#> 40    0.909361587      0
#> 41    0.018197048      0
#> 42   -0.911285290      0
#> 43   -1.108431090      0
#> 44   -0.190431589      0
#> 45    0.523345861      0
#> 46    2.336614093      0
#> 47   -0.794803987      0
#> 48    0.781572764      0
#> 49   -0.773325063      0
#> 50    0.203206894      0
#> 51    0.803502297      0
#> 52   -0.191929007      0
#> 53   -1.164145417      0
#> 54   -1.383629746      0
#> 55    2.369044355      0
#> 56    2.054465802      0
#> 57    1.588577399      0
#> 58    1.011115169      0
#> 59    1.380590569      0
#> 60    1.140580908      0
#> 61   -0.906927304      0
#> 62   -0.381421849      0
#> 63   -0.157061147      0
#> 64    1.222256938      0
#> 65   -0.764319631      0
#> 66   -0.419654012      0
#> 67    1.238377098      0
#> 68   -0.651994310      0
#> 69   -0.413943981      0
#> 70    0.538559892      0
#> 71   -0.351146949      0
#> 72   -1.777099443      0
#> 73    0.347927903      0
#> 74    0.650793645      0
#> 75   -1.251754184      0
#> 76    1.070746093      0
#> 77    1.021796733      0
#> 78   -1.148851351      0
#> 79   -0.321755388      0
#> 80   -1.910148123      0
#> 81    0.454776823      0
#> 82    0.143147724      0
#> 83   -0.156544240      0
#> 84   -0.901449263      0
#> 85   -0.534913086      0
#> 86    0.943233672      0
#> 87   -0.518020886      0
#> 88    0.045764837      0
#> 89   -0.429335115      0
#> 90   -2.041341394      0
#> 91   -0.253962443      0
#> 92   -1.924839042      0
#> 93    0.811670744      0
#> 94    0.601472419      0
#> 95   -0.424641481      0
#> 96    0.662112280      0
#> 97   -0.755998050      0
#> 98   -0.631717557      0
#> 99   -1.934963915      0
#> 100  -0.458318271      0
#> 101  -1.650151374      0
#> 102   0.065936183      0
#> 103  -0.627821830      0
#> 104   0.171886949      0
#> 105  -0.926968834      0
#> 106   0.933915150      0
#> 107   0.400185269      0
#> 108  -0.220613186      0
#> 109  -0.592764494      0
#> 110   1.637442838      0
#> 111   1.792526155      0
#> 112  -1.227113319      0
#> 113   1.559639426      0
#> 114  -2.615778040      0
#> 115  -0.348973047      0
#> 116   1.028639943      0
#> 117   0.039170969      0
#> 118  -0.129837210      0
#> 119  -1.484266146      0
#> 120  -0.327345294      0
#> 121  -0.681920093      0
#> 122  -0.227508822      0
#> 123   0.457090635      0
#> 124  -0.834919358      0
#> 125  -0.330608018      0
#> 126  -0.224120294      0
#> 127   0.160602260      0
#> 128   0.463887675      0
#> 129   0.141519974      0
#> 130   0.586558871      0
#> 131  -1.818440324      0
#> 132  -0.908469865      0
#> 133   0.228954013      0
#> 134  -1.065396483      0
#> 135  -1.544335762      0
#> 136   0.535867035      0
#> 137  -2.538980808      0
#> 138  -0.662580588      0
#> 139   0.230207145      0
#> 140   0.108754802      0
#> 141  -0.391049774      0
#> 142   0.361304293      0
#> 143  -0.261744514      0
#> 144   0.925493358      0
#> 145  -1.576389389      0
#> 146  -1.464415356      0
#> 147   0.366708863      0
#> 148   0.083828387      0
#> 149  -1.344313487      0
#> 150  -0.051902191      0
#> 151  -0.666775192      0
#> 152   0.664415077      0
#> 153  -0.632211889      0
#> 154  -0.819797090      0
#> 155   0.011112999      0
#> 156   0.340526734      0
#> 157  -0.425007476      0
#> 158  -2.299494746      0
#> 159   0.138245472      0
#> 160  -0.236983148      0
#> 161   0.131009353      0
#> 162   0.150254566      0
#> 163  -0.515184122      0
#> 164  -0.254084385      0
#> 165   0.479009286      0
#> 166   0.177234708      0
#> 167   0.023112477      0
#> 168   0.549494832      0
#> 169  -0.014488907      0
#> 170  -1.276317969      0
#> 171  -0.734972902      0
#> 172   1.344698850      0
#> 173  -0.557706936      0
#> 174   0.918659364      0
#> 175  -1.120122676      0
#> 176   0.544265557      0
#> 177   0.130627278      0
#> 178  -0.895161924      0
#> 179   1.199224863      0
#> 180   0.780838835      0
#> 181   1.049153447      0
#> 182  -0.601439028      0
#> 183  -0.774849168      0
#> 184  -1.218618623      0
#> 185  -0.030422373      0
#> 186   0.766890939      0
#> 187  -0.209904338      0
#> 188  -0.728958514      0
#> 189   1.897558633      0
#> 190   0.744969094      0
#> 191   0.045608130      0
#> 192  -1.109799361      0
#> 193   0.083333996      0
#> 194   1.008990759      0
#> 195  -0.345826822      0
#> 196   0.443369181      0
#> 197  -0.590282301      0
#> 198  -1.379841972      0
#> 199   0.301147919      0
#> 200   1.244729556      0
#> 201   0.839750472      0
#> 202   0.150935672      0
#> 203  -0.978292860      0
#> 204   0.903471725      0
#> 205  -0.680558617      0
#> 206  -0.591163475      0
#> 207   0.436365760      0
#> 208   0.856002360      0
#> 209   0.611935984      0
#> 210  -0.084923905      0
#> 211   0.151869958      0
#> 212   1.128059218      0
#> 213   0.301186903      0
#> 214   0.638441188      0
#> 215  -1.284760711      0
#> 216  -0.027642416      0
#> 217  -0.097813579      0
#> 218  -0.474608827      0
#> 219  -0.280207795      0
#> 220   1.348354307      0
#> 221  -0.490677550      0
#> 222  -0.671032792      0
#> 223  -0.089349851      0
#> 224  -0.220722381      0
#> 225  -0.857713296      0
#> 226  -2.183365671      0
#> 227   0.743807928      0
#> 228   0.039500623      0
#> 229   0.674595749      0
#> 230   0.316501492      0
#> 231  -1.236561480      0
#> 232  -2.431849048      0
#> 233   0.268162051      0
#> 234   1.046814970      0
#> 235  -1.124267397      0
#> 236   0.023727919      0
#> 237   0.103415342      0
#> 238   1.114262081      0
#> 239  -1.605512375      0
#> 240   1.662898224      0
#> 241  -0.266973376      0
#> 242  -0.009159038      0
#> 243  -0.341202666      0
#> 244   1.239617766      0
#> 245  -0.390997805      0
#> 246   0.639074809      0
#> 247   1.216431191      0
#> 248   0.073162539      0
#> 249   0.216383066      0
#> 250  -0.047090060      0
#> 251   0.101718329      0
#> 252   0.236180283      0
#> 253  -0.638096624      0
#> 254   1.123257869      0
#> 255   2.131759772      0
#> 256   0.836076282      0
#> 257   0.500141065      0
#> 258   1.483797970      0
#> 259   0.997689171      0
#> 260  -1.167765614      0
#> 261  -0.103144625      0
#> 262   2.195034989      0
#> 263  -0.018357244      0
#> 264  -0.311826439      0
#> 265  -2.059634413      0
#> 266   0.109991530      0
#> 267   0.390949588      0
#> 268   0.906935360      0
#> 269  -1.569173100      0
#> 270  -1.052089594      0
#> 271  -0.389267974      0
#> 272  -0.341389565      0
#> 273  -0.317444008      0
#> 274  -0.668562436      0
#> 275   0.573394076      0
#> 276  -1.179511860      0
#> 277   0.344828941      0
#> 278   1.630197207      0
#> 279  -0.681854057      0
#> 280   0.674623251      0
#> 281   0.446213370      0
#> 282   0.188766635      0
#> 283  -0.395748274      0
#> 284  -1.167594310      0
#> 285   0.778627331      0
#> 286  -0.012911110      0
#> 287   1.295252878      0
#> 288   1.083699640      0
#> 289  -0.030949530      0
#> 290  -1.447070308      0
#> 291  -0.695786988      0
#> 292   1.002930015      0
#> 293  -0.206896885      0
#> 294   0.812629231      0
#> 295   2.123161454      0
#> 296   0.472667047      0
#> 297   1.603779984      0
#> 298  -0.961860024      0
#> 299   0.415372333      0
#> 300  -0.429483630      0
#> 301  -1.310024663      0
#> 302  -1.445610685      0
#> 303   0.396647479      0
#> 304   1.241276510      0
#> 305   0.043137166      0
#> 306  -0.918522651      0
#> 307   0.412931068      0
#> 308   0.196735558      0
#> 309   0.126845596      0
#> 310   0.730903842      0
#> 311   0.784939451      0
#> 312  -0.373296261      0
#> 313   0.676155398      0
#> 314  -2.375573029      0
#> 315  -1.755554847      0
#> 316   1.020058162      0
#> 317   1.077269691      0
#> 318  -0.602803716      0
#> 319  -1.110380092      0
#> 320   0.719763778      0
#> 321   0.288011045      0
#> 322   0.427727881      0
#> 323  -1.424527709      0
#> 324   1.289195486      0
#> 325  -1.180975400      0
#> 326   0.009390145      0
#> 327  -0.057161457      0
#> 328  -0.800886983      0
#> 329   1.439598136      0
#> 330   0.678856903      0
#> 331   0.244849987      0
#> 332  -1.609437329      0
#> 333  -0.994088683      0
#> 334   0.372175010      0
#> 335   0.371303332      0
#> 336  -0.392702478      0
#> 337   2.300646862      0
#> 338  -0.027981167      0
#> 339   1.302829273      0
#> 340  -0.911663775      0
#> 341  -0.267300624      0
#> 342  -0.548691893      0
#> 343  -0.512119494      0
#> 344   0.362471861      0
#> 345   1.377915830      0
#> 346   0.578953788      0
#> 347  -0.916529957      0
#> 348   0.127305261      0
#> 349  -0.865152830      0
#> 350  -0.780042760      0
#> 351   0.328591167      0
#> 352  -1.254769626      0
#> 353  -0.011043938      0
#> 354  -0.213765437      0
#> 355  -1.048390779      0
#> 356  -0.523259975      0
#> 357  -0.288129649      0
#> 358  -1.426448203      0
#> 359   1.684081581      0
#> 360  -1.033963840      0
#> 361   0.910756680      0
#> 362   0.283324848      0
#> 363   0.364571016      0
#> 364  -1.021401790      0
#> 365  -0.132872115      0
#> 366   1.143767005      0
#> 367  -0.961796909      0
#> 368   0.807901579      0
#> 369   1.233731653      0
#> 370  -0.545611812      0
#> 371   0.751072273      0
#> 372  -0.184795679      0
#> 373   0.707571832      0
#> 374  -0.627725878      0
#> 375  -0.535421780      0
#> 376  -1.267409241      0
#> 377  -0.156896426      0
#> 378  -0.426744873      0
#> 379   1.035986756      0
#> 380  -1.262888767      0
#> 381   1.687548362      0
#> 382  -0.065324277      0
#> 383  -0.711313435      0
#> 384   0.342240800      0
#> 385  -0.086512871      0
#> 386  -0.717372170      0
#> 387  -1.437088034      0
#> 388  -0.940045757      0
#> 389  -0.247114161      0
#> 390   1.398941157      0
#> 391   0.802330787      0
#> 392   1.171550277      0
#> 393  -0.236423616      0
#> 394  -0.595923755      0
#> 395   0.370725205      0
#> 396  -0.913623977      0
#> 397   1.201161596      0
#> 398  -0.068194050      0
#> 399  -0.299331587      0
#> 400  -0.764588672      0
#> 401   0.715095767      0
#> 402   0.538507003      0
#> 403   1.343517320      0
#> 404   0.751703401      0
#> 405  -1.078235974      0
#> 406   0.174173523      0
#> 407   0.160631377      0
#> 408   0.963761979      0
#> 409  -1.341593926      0
#> 410   1.521878866      0
#> 411   0.128634219      0
#> 412  -0.254287132      0
#> 413  -0.120695556      0
#> 414  -3.272815317      0
#> 415   0.878349361      0
#> 416   0.632698719      0
#> 417  -0.878370596      0
#> 418  -0.993966852      0
#> 419  -0.629239130      0
#> 420   0.004598591      0
#> 421   0.029861802      0
#> 422  -0.210893198      0
#> 423   0.084135568      0
#> 424   0.758598295      0
#> 425  -0.154891965      0
#> 426  -0.651918044      0
#> 427   0.349801239      0
#> 428  -0.690220317      0
#> 429  -0.715972366      0
#> 430   1.185957155      0
#> 431   0.620813557      0
#> 432   1.165652521      0
#> 433   1.685413844      0
#> 434   0.214615090      0
#> 435  -0.999673492      0
#> 436   0.858865372      0
#> 437  -0.559876703      0
#> 438  -0.820499576      0
#> 439   0.226484781      0
#> 440  -1.583353978      0
#> 441   0.570013289      0
#> 442  -1.730875230      0
#> 443   0.540526698      0
#> 444   0.057338494      0
#> 445  -0.662864645      0
#> 446  -0.262355085      0
#> 447  -0.205279812      0
#> 448   0.416520012      0
#> 449   0.574989613      0
#> 450   0.055700869      0
#> 451   0.705701892      0
#> 452  -0.704884777      0
#> 453   0.872041746      0
#> 454   1.311428816      0
#> 455   0.259254561      0
#> 456   1.497128683      0
#> 457  -1.265074493      0
#> 458   0.455665205      0
#> 459   0.891929555      0
#> 460   0.188654678      0
#> 461  -1.855150910      0
#> 462  -0.574553279      0
#> 463  -0.650381840      0
#> 464   1.780477959      0
#> 465  -1.191598389      0
#> 466  -2.044004277      0
#> 467   0.496900894      0
#> 468   2.027665479      0
#> 469   1.149014686      0
#> 470   3.383568311      0
#> 471  -2.684505775      0
#> 472  -0.672129058      0
#> 473  -0.918402605      0
#> 474   0.235019427      0
#> 475   1.156946294      0
#> 476  -0.476939533      0
#> 477  -0.247923875      0
#> 478  -0.830431441      0
#> 479  -0.955139088      0
#> 480  -0.713941229      0
#> 481  -0.760240835      0
#> 482   1.282274770      0
#> 483  -0.264739770      0
#> 484  -2.047610389      0
#> 485   0.352034294      0
#> 486  -1.533028387      0
#> 487   0.561502115      0
#> 488   0.105269914      0
#> 489   0.914486309      0
#> 490   1.602375373      0
#> 491   1.070938543      0
#> 492  -1.173680948      0
#> 493   0.019750009      0
#> 494   1.504952163      0
#> 495  -1.514279414      0
#> 496  -1.003815547      0
#> 497   2.145721327      0
#> 498   1.354364323      0
#> 499  -0.397789228      0
#> 500  -0.870473195      0
#> 501  -0.553574626      0
#> 502  -0.883225757      0
#> 503   0.832499546      0
#> 504  -1.672480886      0
#> 505  -0.284118813      0
#> 506   0.773511750      0
#> 507  -1.637050891      0
#> 508  -0.958734316      0
#> 509   0.314459052      0
#> 510  -0.638571562      0
#> 511  -0.689700456      0
#> 512   0.216675456      0
#> 513   0.687895948      0
#> 514   0.602866180      0
#> 515  -0.027844274      0
#> 516  -1.193328816      0
#> 517  -0.265008477      0
#> 518  -0.670295705      0
#> 519  -0.113637258      0
#> 520  -0.787296175      0
#> 521  -0.702639595      0
#> 522   0.987177970      0
#> 523   0.918447691      0
#> 524  -0.318175661      0
#> 525   1.638508340      0
#> 526  -0.231667068      0
#> 527  -0.571527276      0
#> 528   0.804358161      0
#> 529   0.016832030      0
#> 530   1.571596317      0
#> 531  -1.915904719      0
#> 532   0.684371997      0
#> 533   0.747713258      0
#> 534   1.491788402      0
#> 535   0.417979828      0
#> 536   0.560382508      0
#> 537  -1.567104227      0
#> 538   1.938099797      0
#> 539   0.262207681      0
#> 540  -1.276433145      0
#> 541   1.029249985      0
#> 542   0.105081752      0
#> 543  -1.241281375      0
#> 544   0.318706834      0
#> 545   1.496315288      0
#> 546   1.197430893      0
#> 547   0.249689191      0
#> 548  -0.858218485      0
#> 549  -0.356923937      0
#> 550   1.544500540      0
#> 551   1.148435686      0
#> 552  -1.323196201      0
#> 553   1.432882259      0
#> 554  -0.095198170      0
#> 555  -0.272870853      0
#> 556  -1.481674891      0
#> 557   0.617983055      0
#> 558  -1.425819588      0
#> 559  -0.946199356      0
#> 560  -1.632605557      0
#> 561   1.313210696      0
#> 562   0.201738284      0
#> 563  -0.565297090      0
#> 564   0.091292519      0
#> 565   1.001913628      0
#> 566  -0.847317117      0
#> 567   1.139842402      0
#> 568  -1.155252384      0
#> 569   0.857646769      0
#> 570  -0.897499001      0
#> 571  -0.300245492      0
#> 572  -0.461541389      0
#> 573  -0.703385052      0
#> 574   0.804630445      0
#> 575  -0.015826000      0
#> 576  -0.178171519      0
#> 577  -1.046647532      0
#> 578   0.283941852      0
#> 579  -0.821357915      0
#> 580  -1.944689083      0
#> 581  -0.144664021      0
#> 582   0.575911614      0
#> 583  -1.740301737      0
#> 584  -1.275918594      0
#> 585  -0.226193063      0
#> 586  -0.550676063      0
#> 587  -1.645507406      0
#> 588   0.100230624      0
#> 589  -0.616749280      0
#> 590   1.049917698      0
#> 591   0.344831567      0
#> 592   0.832637309      0
#> 593   0.438591611      0
#> 594   1.557954542      0
#> 595   0.373520475      0
#> 596  -0.785955867      0
#> 597   1.053369821      0
#> 598   0.407937296      0
#> 599   1.134950575      0
#> 600  -1.150169593      0
#> 601  -0.741988322      0
#> 602   1.236397718      0
#> 603  -0.426728573      0
#> 604   0.836867453      0
#> 605   0.121799195      0
#> 606  -1.225946561      0
#> 607  -1.097078300      0
#> 608  -1.196056997      0
#> 609  -0.117050892      0
#> 610   0.749876835      0
#> 611  -0.373582609      0
#> 612   0.189996984      0
#> 613   0.543340453      0
#> 614  -0.088983145      0
#> 615   0.503445676      0
#> 616  -1.045368083      0
#> 617  -0.658281834      0
#> 618   1.235211375      0
#> 619  -0.989260613      0
#> 620   0.944473692      0
#> 621  -1.208762157      0
#> 622  -0.885541246      0
#> 623   0.726932974      0
#> 624  -2.172510753      0
#> 625   0.715587399      0
#> 626   0.805250658      0
#> 627   1.849971100      0
#> 628   0.918162879      0
#> 629  -0.138324651      0
#> 630   0.251517888      0
#> 631   0.420296701      0
#> 632   0.853773269      0
#> 633   0.706377391      0
#> 634   0.215856851      0
#> 635  -0.587372090      0
#> 636   0.597501156      0
#> 637   2.063475015      0
#> 638  -0.372163589      0
#> 639  -0.707082654      0
#> 640  -2.215304698      0
#> 641   0.803350499      0
#> 642   1.417392896      0
#> 643  -0.003566627      0
#> 644   0.366081268      0
#> 645  -0.847082188      0
#> 646   0.658516300      0
#> 647   1.471322636      0
#> 648  -0.988731339      0
#> 649  -0.432404138      0
#> 650   0.997285665      0
#> 651  -0.767563454      0
#> 652   1.542206395      0
#> 653  -0.304810669      0
#> 654  -1.179450795      0
#> 655  -0.951621163      0
#> 656  -0.365501925      0
#> 657  -0.582611556      0
#> 658   0.815140966      0
#> 659  -0.111008736      0
#> 660   0.241445178      0
#> 661  -0.248862873      0
#> 662  -1.056124547      0
#> 663   1.411426301      0
#> 664  -0.479584441      0
#> 665  -0.235194371      0
#> 666  -0.375373311      0
#> 667   0.161235887      0
#> 668  -1.262099442      0
#> 669  -0.853044303      0
#> 670  -1.499598133      0
#> 671  -2.195377747      0
#> 672   0.847359124      0
#> 673  -1.384230083      0
#> 674   0.810083590      0
#> 675   0.665557019      0
#> 676   0.932274174      0
#> 677   1.950157024      0
#> 678  -1.866151276      0
#> 679   1.182213379      0
#> 680   1.148234358      0
#> 681   0.160722099      0
#> 682   1.174427068      0
#> 683  -1.242597311      0
#> 684   0.772984243      0
#> 685   0.522831053      0
#> 686  -1.484244549      0
#> 687   0.015606743      0
#> 688   0.972883361      0
#> 689  -0.602507673      0
#> 690   2.538629318      0
#> 691  -0.484092434      0
#> 692   0.440229423      0
#> 693  -0.139420998      0
#> 694  -0.241063018      0
#> 695   1.334011777      0
#> 696  -0.246860434      0
#> 697  -0.928257442      0
#> 698  -1.087849949      0
#> 699   0.341750528      0
#> 700   1.771403302      0
#> 701   1.102211533      0
#> 702   0.561695528      0
#> 703  -1.064770548      0
#> 704   0.967316815      0
#> 705  -0.165220680      0
#> 706  -0.977851236      0
#> 707   0.343076107      0
#> 708  -0.469361787      0
#> 709   0.265994541      0
#> 710   0.582787016      0
#> 711  -0.626598585      0
#> 712   0.764699017      0
#> 713  -0.390494243      0
#> 714   0.121198424      0
#> 715   0.689028973      0
#> 716  -0.199552051      0
#> 717   0.207807999      0
#> 718  -0.796342364      0
#> 719  -0.559278122      0
#> 720   0.803595461      0
#> 721   1.022302521      0
#> 722   0.034728687      0
#> 723   0.772763307      0
#> 724  -0.374330752      0
#> 725   2.111908015      0
#> 726   0.554195738      0
#> 727  -2.814355625      0
#> 728  -2.423824267      0
#> 729  -0.208756539      0
#> 730   0.447180217      0
#> 731   2.498632684      0
#> 732  -3.103099936      0
#> 733  -0.003568373      0
#> 734  -1.722610428      0
#> 735   0.458123306      0
#> 736  -1.294327517      0
#> 737  -0.388525130      0
#> 738   0.360022026      0
#> 739   0.028889023      0
#> 740  -0.684793046      0
#> 741  -0.826574477      0
#> 742  -1.050115748      0
#> 743   2.830472980      0
#> 744  -0.373418907      0
#> 745  -1.374242325      0
#> 746  -0.897664804      0
#> 747   0.119072799      0
#> 748  -0.231343699      0
#> 749  -1.547046512      0
#> 750  -0.222320688      0
#> 751  -2.714434121      0
#> 752  -1.298698605      0
#> 753   1.502546958      0
#> 754  -0.033820734      0
#> 755   1.508531204      0
#> 756   1.123206680      0
#> 757  -0.406386116      0
#> 758   0.071791847      0
#> 759   0.354848869      0
#> 760   0.067693613      0
#> 761  -0.932198882      0
#> 762   1.763753173      0
#> 763   2.290054537      0
#> 764  -0.424394172      0
#> 765   1.220296488      0
#> 766  -0.219180105      0
#> 767  -0.284988060      0
#> 768  -0.520747872      0
#> 769  -0.445975773      0
#> 770   0.109422163      0
#> 771  -0.693418340      0
#> 772   1.357918999      0
#> 773   0.293552968      0
#> 774   1.031903517      0
#> 775   1.097980813      0
#> 776   0.618765016      0
#> 777   0.096724867      0
#> 778  -1.141904074      0
#> 779   1.125287049      0
#> 780   1.024523337      0
#> 781  -0.891142443      0
#> 782   0.037944538      0
#> 783  -1.320086782      0
#> 784   1.096434406      0
#> 785   0.985108236      0
#> 786  -0.284277234      0
#> 787  -0.493260080      0
#> 788  -1.186669273      0
#> 789  -0.163270247      0
#> 790   0.366980091      0
#> 791   1.038540038      0
#> 792  -0.433606243      0
#> 793   1.616233076      0
#> 794   1.675394054      0
#> 795   0.056226920      0
#> 796  -1.372036815      0
#> 797  -1.557569429      0
#> 798  -1.409149547      0
#> 799  -0.653955183      0
#> 800   0.749993712      0
#> 801  -0.700591289      0
#> 802   0.070039203      0
#> 803  -1.118829514      0
#> 804  -1.575444229      0
#> 805  -1.672259388      0
#> 806  -0.088578744      0
#> 807  -1.241101172      0
#> 808   0.566991440      0
#> 809  -0.203307715      0
#> 810  -1.435950922      0
#> 811  -1.234151218      0
#> 812   0.111664968      0
#> 813   0.861284587      0
#> 814  -0.216735318      0
#> 815  -0.289715892      0
#> 816   0.709478815      0
#> 817   1.010510307      0
#> 818   1.341015521      0
#> 819   0.035848884      0
#> 820  -1.009399980      0
#> 821  -1.014623848      0
#> 822  -2.234833615      0
#> 823  -0.083747309      0
#> 824   0.614050930      0
#> 825   0.476419188      0
#> 826   0.693192066      0
#> 827  -1.072349130      0
#> 828   0.335955267      0
#> 829  -0.741166630      0
#> 830   0.335958835      0
#> 831   1.150662262      0
#> 832  -0.393216214      0
#> 833  -0.678690474      0
#> 834  -0.255937607      0
#> 835   0.345272210      0
#> 836   0.860563189      0
#> 837   0.478729511      0
#> 838  -0.296679965      0
#> 839   1.560716139      0
#> 840   0.324941342      0
#> 841  -1.067832862      0
#> 842   0.508655373      0
#> 843   0.596530472      0
#> 844   0.866301438      0
#> 845  -0.330468009      0
#> 846   0.415437905      0
#> 847  -0.447781394      0
#> 848   0.909724389      0
#> 849   0.743282551      0
#> 850   0.455238703      0
#> 851  -1.450086022      0
#> 852  -1.301907186      0
#> 853  -0.539953537      0
#> 854  -0.956947605      0
#> 855   1.054429952      0
#> 856  -1.199869603      0
#> 857   0.151495483      0
#> 858  -0.372720299      0
#> 859  -1.039839809      0
#> 860   1.695780590      0
#> 861   0.758480226      0
#> 862   0.149354137      0
#> 863  -1.113278532      0
#> 864  -0.999072765      0
#> 865  -1.569015046      0
#> 866   0.084702045      0
#> 867  -0.432139191      0
#> 868  -0.144262094      0
#> 869  -1.989862246      0
#> 870   0.978244220      0
#> 871   0.316543052      0
#> 872   0.273409708      0
#> 873   0.473747026      0
#> 874  -0.438821208      0
#> 875   1.576643765      0
#> 876  -1.733974131      0
#> 877  -0.304793608      0
#> 878  -0.211027246      0
#> 879  -0.512959577      0
#> 880  -1.022032991      0
#> 881  -0.565801151      0
#> 882   1.275369821      0
#> 883   1.072683027      0
#> 884   0.801379004      0
#> 885   0.888980656      0
#> 886   1.280659492      0
#> 887   0.745667054      0
#> 888   0.181987885      0
#> 889  -0.848577181      0
#> 890   0.598862509      0
#> 891   0.212016933      0
#> 892   0.034469809      0
#> 893  -0.539402344      0
#> 894  -0.052660741      0
#> 895  -0.761802710      0
#> 896  -1.224276353      0
#> 897   1.012438126      0
#> 898  -0.170191315      1
#> 899   0.752284654      1
#> 900   2.058722552      1
#> 901   0.237944608      1
#> 902  -0.294867352      1
#> 903   1.634340755      1
#> 904   1.160826576      1
#> 905   1.221734031      1
#> 906  -0.046599074      1
#> 907   0.909122311      1
#> 908   0.676269092      1
#> 909   3.265963465      1
#> 910  -0.486280879      1
#> 911  -0.524824299      1
#> 912   0.596942845      1
#> 913   1.295881816      1
#> 914   1.207623433      1
#> 915   0.673631705      1
#> 916  -1.301409216      1
#> 917   0.933243762      1
#> 918  -0.638321566      1
#> 919   2.615744931      1
#> 920   0.369694787      1
#> 921  -0.847674884      1
#> 922   0.969898837      1
#> 923   0.415695641      1
#> 924  -0.441313914      1
#> 925   0.484126752      1
#> 926   0.746319080      1
#> 927   1.426399468      1
#> 928  -0.031933699      1
#> 929  -0.710638276      1
#> 930   0.177600644      1
#> 931   1.282649974      1
#> 932   0.953242447      1
#> 933  -0.770502130      1
#> 934   0.928093885      1
#> 935   2.217878995      1
#> 936   0.541005069      1
#> 937  -0.378292548      1
#> 938   1.182406686      1
#> 939   0.723261758      1
#> 940   2.293223056      1
#> 941   1.910704244      1
#> 942   1.081073685      1
#> 943  -0.712634051      1
#> 944   1.086841005      1
#> 945   0.796146282      1
#> 946   1.426906857      1
#> 947   1.405676840      1
#> 948   0.846771992      1
#> 949   1.892118105      1
#> 950  -0.031203210      1
#> 951   0.897043401      1
#> 952   1.440509121      1
#> 953   0.175597014      1
#> 954  -0.060110489      1
#> 955   0.745453351      1
#> 956   0.556549059      1
#> 957   1.044676284      1
#> 958  -0.847502114      1
#> 959   2.341196580      1
#> 960   2.724520477      1
#> 961   1.449009552      1
#> 962   1.051141356      1
#> 963   0.337907442      1
#> 964   0.144786317      1
#> 965   0.099722854      1
#> 966   0.386439583      1
#> 967   0.507028848      1
#> 968   0.465452311      1
#> 969   0.629700859      1
#> 970   1.322678602      1
#> 971  -0.041262513      1
#> 972  -0.252695738      1
#> 973   0.612419145      1
#> 974   0.524599110      1
#> 975   2.784044531      1
#> 976   0.051272869      1
#> 977   0.261899440      1
#> 978   0.484126758      1
#> 979   1.717593586      1
#> 980   1.288117540      1
#> 981   1.282417167      1
#> 982   1.179218081      1
#> 983   2.312385143      1
#> 984   0.143805142      1
#> 985  -0.347962387      1
#> 986   1.752906794      1
#> 987   2.721343327      1
#> 988   0.176951203      1
#> 989   1.636240798      1
#> 990   1.051778434      1
#> 991  -0.426814685      1
#> 992   1.055137303      1
#> 993   0.366863600      1
#> 994   1.206117812      1
#> 995  -0.087161094      1
#> 996   0.841261701      1
#> 997   0.906548232      1
#> 998   1.165844209      1
#> 999   0.893623078      1
#> 1000 -0.463582555      1