Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -1.1135165194      0
#> 2    -2.0020502067      0
#> 3    -3.1662348452      0
#> 4     1.0042966545      0
#> 5     0.8146331363      0
#> 6     1.1970965548      0
#> 7    -0.5346573505      0
#> 8     1.0398336366      0
#> 9    -0.7058577909      0
#> 10    2.6203755384      0
#> 11   -1.4620800978      0
#> 12   -0.8481363174      0
#> 13    0.8200081098      0
#> 14   -0.1676702080      0
#> 15   -0.2002603944      0
#> 16    0.0129827234      0
#> 17   -0.6462218972      0
#> 18   -0.3826824845      0
#> 19   -1.5376559077      0
#> 20   -0.7743489768      0
#> 21   -0.8210671288      0
#> 22   -1.0718018274      0
#> 23    0.9558961533      0
#> 24    0.2081719127      0
#> 25   -0.1860088936      0
#> 26   -1.6590939413      0
#> 27   -0.8054314341      0
#> 28   -1.2769503237      0
#> 29    1.5350857783      0
#> 30    1.0969357724      0
#> 31   -0.0255720489      0
#> 32   -0.4275786120      0
#> 33    1.7434002756      0
#> 34    0.0808552113      0
#> 35    1.3527791847      0
#> 36    0.6476786495      0
#> 37   -1.4034690880      0
#> 38   -0.2692158704      0
#> 39   -1.1006486807      0
#> 40   -1.3960609369      0
#> 41   -0.8460080951      0
#> 42    0.5612131775      0
#> 43   -1.3469388803      0
#> 44   -2.2688836145      0
#> 45   -0.0130253875      0
#> 46   -1.6402473887      0
#> 47   -1.0436337201      0
#> 48    0.3678019688      0
#> 49    0.8800967431      0
#> 50    1.5308495615      0
#> 51    0.2073671515      0
#> 52   -0.7527181164      0
#> 53    0.5929083341      0
#> 54   -0.7582323318      0
#> 55   -0.9205104557      0
#> 56    0.7182893838      0
#> 57    1.1277345739      0
#> 58    1.3311530884      0
#> 59   -1.4632189642      0
#> 60   -1.1983653378      0
#> 61    1.0256501407      0
#> 62   -1.1897089706      0
#> 63   -1.0519825579      0
#> 64    0.0187250139      0
#> 65    0.5997464904      0
#> 66   -0.7767191652      0
#> 67    0.9146005951      0
#> 68   -0.3715247359      0
#> 69    0.0858561845      0
#> 70   -0.0258768197      0
#> 71    1.0122815189      0
#> 72   -0.7589979100      0
#> 73   -0.9953724664      0
#> 74    0.6539196179      0
#> 75    0.5395227369      0
#> 76    0.9091593662      0
#> 77    0.5954471794      0
#> 78    1.3277456325      0
#> 79   -0.7040479923      0
#> 80   -1.0653769541      0
#> 81    0.1154239852      0
#> 82    1.4583298791      0
#> 83    0.8254103800      0
#> 84    0.2806690348      0
#> 85   -0.5999076792      0
#> 86   -0.4991180279      0
#> 87    1.2718323904      0
#> 88    0.2785849878      0
#> 89    0.0114807486      0
#> 90   -0.1204557567      0
#> 91    1.7201234487      0
#> 92   -0.0223622479      0
#> 93    0.6768719343      0
#> 94    1.1880641494      0
#> 95    1.8794032769      0
#> 96    0.0875349118      0
#> 97   -0.3530082259      0
#> 98    1.0898195058      0
#> 99   -0.2030118360      0
#> 100  -0.3279785244      0
#> 101  -1.2403466151      0
#> 102   0.4118186964      0
#> 103  -0.7097695529      0
#> 104  -0.6505946792      0
#> 105   2.9496306396      0
#> 106  -0.1407260228      0
#> 107  -1.2822586868      0
#> 108   0.2596594511      0
#> 109   1.7334061662      0
#> 110   0.7951490793      0
#> 111  -2.2841795888      0
#> 112   0.8316226736      0
#> 113  -0.3416785518      0
#> 114  -1.0306448111      0
#> 115  -0.4325453661      0
#> 116   1.6662396299      0
#> 117  -0.7412365871      0
#> 118  -0.5126470083      0
#> 119  -0.2488552265      0
#> 120  -0.6986722140      0
#> 121   1.1685616626      0
#> 122  -0.6982024670      0
#> 123  -0.5282686752      0
#> 124  -0.2850346262      0
#> 125  -0.8315791595      0
#> 126  -0.6409633221      0
#> 127   0.3201980086      0
#> 128  -1.2923428921      0
#> 129   1.6901561622      0
#> 130   1.2284612050      0
#> 131   0.4875044672      0
#> 132  -0.6468244536      0
#> 133  -0.7326158880      0
#> 134  -0.3193907570      0
#> 135  -1.0292739924      0
#> 136   0.1770275321      0
#> 137  -1.2520690058      0
#> 138   0.7587261317      0
#> 139  -0.8272864762      0
#> 140   1.6685650770      0
#> 141   0.9601481388      0
#> 142  -1.4636167719      0
#> 143  -0.1765807033      0
#> 144   0.4874747447      0
#> 145  -1.3062963807      0
#> 146   0.6692510335      0
#> 147  -2.4604466591      0
#> 148   1.1135001791      0
#> 149   0.3460725534      0
#> 150  -0.0786806944      0
#> 151  -1.0658680118      0
#> 152  -0.4783509384      0
#> 153  -1.3334652033      0
#> 154  -1.2340641270      0
#> 155  -0.5861689414      0
#> 156   0.7208332802      0
#> 157   0.8895087879      0
#> 158  -0.7922331532      0
#> 159   1.5254888614      0
#> 160  -1.0114643449      0
#> 161  -0.4658819073      0
#> 162  -1.4390801517      0
#> 163  -1.1129331486      0
#> 164  -0.5149050017      0
#> 165   0.6116628322      0
#> 166  -1.2188372046      0
#> 167   1.6557354078      0
#> 168   0.8042653316      0
#> 169  -0.4823616186      0
#> 170  -1.2674382296      0
#> 171  -0.2057532896      0
#> 172   0.3348171403      0
#> 173   0.5381442588      0
#> 174  -1.0601820706      0
#> 175  -0.1311286387      0
#> 176   0.2231214917      0
#> 177   0.4757025764      0
#> 178  -0.8967260415      0
#> 179   0.0011861400      0
#> 180  -1.0480331289      0
#> 181  -0.1890132710      0
#> 182   1.4676699496      0
#> 183  -1.3298707067      0
#> 184  -0.6799898954      0
#> 185   0.6784775359      0
#> 186   0.6796947207      0
#> 187   0.1897212688      0
#> 188  -0.8617037301      0
#> 189  -0.9208002703      0
#> 190   0.0370138688      0
#> 191   1.4423601782      0
#> 192  -1.0612863119      0
#> 193   0.9083231962      0
#> 194  -0.8703352643      0
#> 195  -0.9378887244      0
#> 196   0.3780608802      0
#> 197   1.2465459756      0
#> 198   0.0380212991      0
#> 199  -1.3612681058      0
#> 200   0.9391757763      0
#> 201  -0.1564782589      0
#> 202   2.4865724651      0
#> 203  -2.0480765878      0
#> 204  -1.9692076545      0
#> 205  -1.2814703137      0
#> 206   0.6989793807      0
#> 207   0.8625902063      0
#> 208  -1.5652524924      0
#> 209  -0.1335586586      0
#> 210   0.4208829814      0
#> 211  -1.1692506743      0
#> 212  -0.3629323807      0
#> 213  -0.8012905155      0
#> 214  -0.2340372593      0
#> 215   1.6584841658      0
#> 216   0.4000157562      0
#> 217   0.7334371992      0
#> 218  -0.0678832167      0
#> 219   0.2273960151      0
#> 220  -0.4287573513      0
#> 221   0.2901387521      0
#> 222   0.4920535233      0
#> 223  -0.2834579517      0
#> 224   1.5319219283      0
#> 225  -0.3762819634      0
#> 226   0.4446232493      0
#> 227  -1.1524387261      0
#> 228   0.0365121324      0
#> 229   1.2849094731      0
#> 230   1.4498100087      0
#> 231   0.8180556235      0
#> 232  -0.0951443356      0
#> 233  -1.2819760105      0
#> 234   0.8028242412      0
#> 235  -1.3528763461      0
#> 236   1.6913746172      0
#> 237  -1.1989449876      0
#> 238  -0.6723708743      0
#> 239   0.8683748307      0
#> 240  -0.0744282203      0
#> 241   0.8794666174      0
#> 242   0.5867754126      0
#> 243  -0.9991189794      0
#> 244   1.2659370861      0
#> 245  -1.9693688231      0
#> 246  -0.6599080849      0
#> 247  -1.2913415004      0
#> 248   1.0899437366      0
#> 249   1.1585795373      0
#> 250   1.1276855565      0
#> 251  -0.6205055861      0
#> 252  -1.1051536768      0
#> 253  -0.3793859178      0
#> 254   0.6080574850      0
#> 255   0.0481544704      0
#> 256   0.5541037279      0
#> 257  -0.4704247252      0
#> 258   0.4511597221      0
#> 259   0.6681106611      0
#> 260   0.6119529826      0
#> 261  -0.5574186462      0
#> 262  -1.2289029603      0
#> 263   0.5523975873      0
#> 264   1.0934649167      0
#> 265  -1.0668552169      0
#> 266  -2.0550857857      0
#> 267   1.6797078931      0
#> 268  -1.4903132028      0
#> 269  -0.4874087708      0
#> 270  -0.6526346080      0
#> 271  -0.2209593327      0
#> 272  -1.6149330395      0
#> 273   1.9437585230      0
#> 274  -0.5582456322      0
#> 275   0.1637803393      0
#> 276  -0.1299428138      0
#> 277   0.2581542647      0
#> 278  -1.1883889394      0
#> 279   0.5485460280      0
#> 280   0.5167711985      0
#> 281   1.3924295681      0
#> 282   0.0735581455      0
#> 283  -1.6665333976      0
#> 284   0.8044155812      0
#> 285  -0.6029704650      0
#> 286  -1.2159891504      0
#> 287  -0.1020365109      0
#> 288   0.0281473798      0
#> 289   2.3483398297      0
#> 290   0.7658008688      0
#> 291  -0.3247236474      0
#> 292   0.4401308331      0
#> 293  -1.4149525803      0
#> 294  -0.1972713777      0
#> 295  -0.4059153061      0
#> 296   1.6391356988      0
#> 297  -0.2518974034      0
#> 298   1.6836837808      0
#> 299  -2.0340003144      0
#> 300  -0.8579465805      0
#> 301   0.8179449637      0
#> 302  -0.9780515064      0
#> 303  -0.5660592672      0
#> 304  -0.3041978298      0
#> 305   0.7802373330      0
#> 306  -1.1495636617      0
#> 307  -1.1300835968      0
#> 308  -0.5193620697      0
#> 309   0.6991980881      0
#> 310  -1.3771431240      0
#> 311   0.2020367694      0
#> 312  -0.0555751922      0
#> 313   0.0437536333      0
#> 314   0.0286701860      0
#> 315   1.5944591093      0
#> 316   2.2234163160      0
#> 317   1.4845087277      0
#> 318  -1.0567252136      0
#> 319  -0.0849506510      0
#> 320  -0.0539646685      0
#> 321  -2.1472025900      0
#> 322   0.4071967815      0
#> 323   0.8671437974      0
#> 324  -0.1161606582      0
#> 325   0.4007905204      0
#> 326   0.0433940185      0
#> 327  -1.0566559488      0
#> 328   0.4311180536      0
#> 329  -1.8761308563      0
#> 330  -0.2932116041      0
#> 331  -2.0548357236      0
#> 332  -0.2800456185      0
#> 333   0.5467400762      0
#> 334  -1.1744201512      0
#> 335   0.9597549939      0
#> 336  -1.2109729677      0
#> 337  -0.3270699824      0
#> 338   1.2476113254      0
#> 339  -0.3342360368      0
#> 340   0.6363176842      0
#> 341  -1.3480966919      0
#> 342  -0.0928159271      0
#> 343   1.1541257753      0
#> 344   0.5788493392      0
#> 345  -1.1483091968      0
#> 346  -0.0232256719      0
#> 347   2.5942852703      0
#> 348   0.2956537035      0
#> 349   0.3381669489      0
#> 350   1.0514884769      0
#> 351  -0.1904334758      0
#> 352  -1.7016199357      0
#> 353  -1.3560456938      0
#> 354   0.4289653643      0
#> 355  -0.0141510074      0
#> 356   0.7358293998      0
#> 357  -0.0742748303      0
#> 358  -0.3061126265      0
#> 359  -0.3724766783      0
#> 360  -0.8928831142      0
#> 361   0.5947843573      0
#> 362  -0.0408411651      0
#> 363   1.1717827633      0
#> 364  -1.5601488608      0
#> 365  -1.6475363060      0
#> 366   0.9104165178      0
#> 367  -0.3028963538      0
#> 368   0.7563160658      0
#> 369  -0.5515484509      0
#> 370   0.9194473407      0
#> 371  -0.1865134283      0
#> 372  -0.2667202485      0
#> 373   0.9079565282      0
#> 374  -0.1201670975      0
#> 375  -0.8936083896      0
#> 376  -0.7215708750      0
#> 377   1.2435186060      0
#> 378  -0.5279239229      0
#> 379   0.7803218767      0
#> 380  -0.3943488882      0
#> 381  -1.2289846173      0
#> 382   1.1621681620      0
#> 383  -1.1529962051      0
#> 384   0.3990936703      0
#> 385  -0.4545066638      0
#> 386   1.8348186965      0
#> 387   0.9054909380      0
#> 388   1.1212980853      0
#> 389   0.5333503456      0
#> 390   2.0769635190      0
#> 391  -0.1886248245      0
#> 392   0.1758953774      0
#> 393   1.7674759386      0
#> 394   0.0906697331      0
#> 395  -1.7585947125      0
#> 396  -0.4074560172      0
#> 397   0.2240155268      0
#> 398   1.3328407044      0
#> 399  -0.7123362177      0
#> 400   1.5441397238      0
#> 401   1.4918811659      0
#> 402   0.0531429831      0
#> 403   1.0962338961      0
#> 404  -1.2093226925      0
#> 405   0.6299500314      0
#> 406  -1.3332106428      0
#> 407   0.1453392554      0
#> 408   0.5442485297      0
#> 409   1.1375967701      0
#> 410   0.4021179528      0
#> 411   0.6819264468      0
#> 412  -0.8433773522      0
#> 413   0.3993346229      0
#> 414   0.5597795329      0
#> 415   1.4176035045      0
#> 416   1.1605154356      0
#> 417  -0.0181794590      0
#> 418   1.9557324250      0
#> 419  -0.5323181727      0
#> 420  -1.0553972422      0
#> 421   1.2806022827      0
#> 422   0.2858241486      0
#> 423   0.4025051502      0
#> 424   0.8788263266      0
#> 425  -0.0436893191      0
#> 426  -1.4083836261      0
#> 427  -0.4950280211      0
#> 428  -0.5496158238      0
#> 429   1.0520828259      0
#> 430  -0.8984638604      0
#> 431  -1.3049509349      0
#> 432   1.6270266967      0
#> 433  -0.1078849115      0
#> 434   0.8331501236      0
#> 435  -1.7470444734      0
#> 436   0.9906863372      0
#> 437  -0.6142979459      0
#> 438   0.8977490819      0
#> 439  -2.2678816313      0
#> 440  -0.2265596047      0
#> 441  -0.9906865580      0
#> 442  -0.0179340136      0
#> 443   0.4524197695      0
#> 444  -0.1056910159      0
#> 445   0.2355202433      0
#> 446   1.2991882273      0
#> 447   0.1296609669      0
#> 448  -1.5269245691      0
#> 449   0.7885537216      0
#> 450   0.3407523300      0
#> 451  -0.0239046857      0
#> 452  -0.7595113721      0
#> 453  -0.0110159287      0
#> 454  -0.1082384796      0
#> 455   1.0635745333      0
#> 456  -1.8881048114      0
#> 457   0.8272052750      0
#> 458   0.2041482981      0
#> 459  -2.0609880392      0
#> 460  -0.4823284261      0
#> 461  -1.7118282123      0
#> 462   0.6138428885      0
#> 463   0.2715641652      0
#> 464   0.1759713855      0
#> 465  -1.0118794398      0
#> 466   0.6188693109      0
#> 467   1.1870735332      0
#> 468  -0.3242722326      0
#> 469  -0.8913823537      0
#> 470   0.5036022040      0
#> 471   0.0748009834      0
#> 472   1.3166199179      0
#> 473  -1.1526761431      0
#> 474   0.5512713798      0
#> 475   0.6563525085      0
#> 476   1.2306938506      0
#> 477  -0.4369860013      0
#> 478  -1.2913013128      0
#> 479  -0.9306887814      0
#> 480  -2.1091295710      0
#> 481  -0.6148305074      0
#> 482  -0.3729302312      0
#> 483   1.6817175811      0
#> 484  -1.3342929661      0
#> 485   0.1127442758      0
#> 486  -0.3250695123      0
#> 487  -0.8830735550      0
#> 488   0.4749809891      0
#> 489   0.9730065470      0
#> 490   0.3153397546      0
#> 491   1.0391831020      0
#> 492  -0.3417056286      0
#> 493   0.3866173200      0
#> 494   1.3523490604      0
#> 495   0.4940583512      0
#> 496  -0.5726181898      0
#> 497   1.1309737929      0
#> 498   0.0882279915      0
#> 499  -0.6525869638      0
#> 500  -1.9531314988      0
#> 501  -0.7131628465      0
#> 502  -0.0775498650      0
#> 503  -2.0223459595      0
#> 504  -0.7596559126      0
#> 505  -0.1048408023      0
#> 506  -0.7176180850      0
#> 507   0.6602566246      0
#> 508   1.0817566728      0
#> 509  -1.1087839941      0
#> 510  -0.1658858846      0
#> 511  -0.7272993175      0
#> 512  -0.6093032028      0
#> 513  -0.5260445631      0
#> 514  -0.4761157732      0
#> 515  -0.7444520444      0
#> 516   0.2885528208      0
#> 517  -1.2197360348      0
#> 518  -1.7118754213      0
#> 519   0.4137076427      0
#> 520   0.6576407608      0
#> 521  -1.2067614231      0
#> 522   0.6766079554      0
#> 523   1.0773963622      0
#> 524   1.1294909304      0
#> 525   0.1613457977      0
#> 526  -1.5305055987      0
#> 527  -0.9274396430      0
#> 528   0.9743505010      0
#> 529   0.7367777491      0
#> 530   0.4140127007      0
#> 531  -1.9089246393      0
#> 532   0.6951107253      0
#> 533  -0.2507207347      0
#> 534   1.6194148297      0
#> 535   0.7801838890      0
#> 536  -0.1718846565      0
#> 537   1.4225308847      0
#> 538   0.1973169942      0
#> 539   0.3699612726      0
#> 540   0.4469935654      0
#> 541   0.2353578982      0
#> 542   1.2021897211      0
#> 543  -1.3099766375      0
#> 544  -1.4942029440      0
#> 545  -0.7613465444      0
#> 546  -0.1836482074      0
#> 547  -0.0033842208      0
#> 548   2.2067303477      0
#> 549  -0.1924786815      0
#> 550   0.2771428101      0
#> 551   0.9933532799      0
#> 552   0.3739863069      0
#> 553   0.1790029537      0
#> 554   0.2976486494      0
#> 555   0.7689503561      0
#> 556   1.7086628938      0
#> 557   0.2035381603      0
#> 558  -0.3517439074      0
#> 559  -0.4581253248      0
#> 560  -0.2333032861      0
#> 561   1.8302558564      0
#> 562   0.7408411799      0
#> 563  -0.6614506747      0
#> 564   0.8792584199      0
#> 565   0.4684949698      0
#> 566   0.0007231239      0
#> 567  -0.0327364743      0
#> 568  -1.2020765588      0
#> 569  -1.4907152551      0
#> 570   0.0936176569      0
#> 571  -3.0108875180      0
#> 572  -0.5911175963      0
#> 573   0.5728189310      0
#> 574  -0.7566067551      0
#> 575  -0.4104157107      0
#> 576   0.0401219525      0
#> 577   0.7002477442      0
#> 578  -0.4954281272      0
#> 579   1.2685998838      0
#> 580  -2.6110202748      0
#> 581  -1.7930141978      0
#> 582   1.0611925111      0
#> 583  -1.7781243315      0
#> 584  -0.0474986175      0
#> 585  -2.3136881009      0
#> 586  -0.4589674949      0
#> 587  -0.6854556064      0
#> 588  -1.9257754958      0
#> 589   0.6012960522      0
#> 590   0.4167720056      0
#> 591  -0.4609936149      0
#> 592   0.3765009146      0
#> 593  -0.4903937111      0
#> 594  -0.0665785888      0
#> 595  -0.4076608255      0
#> 596   1.0069019855      0
#> 597   0.3386988337      0
#> 598  -0.0932225212      0
#> 599   0.4620287210      0
#> 600   0.4180835970      0
#> 601  -1.8775018726      0
#> 602   3.3110114702      0
#> 603  -0.5092225260      0
#> 604   0.5096794138      0
#> 605  -0.4468065187      0
#> 606   0.6598353337      0
#> 607  -1.7062532224      0
#> 608  -1.5198194420      0
#> 609   0.7738379092      0
#> 610  -1.5065439923      0
#> 611  -1.4058229256      0
#> 612  -0.9065645472      0
#> 613  -0.6980219730      0
#> 614  -0.2690699100      0
#> 615   0.9241269403      0
#> 616  -0.9345169606      0
#> 617   0.8377286836      0
#> 618  -2.5123563333      0
#> 619   0.4217799110      0
#> 620   1.2360695765      0
#> 621   0.6000239920      0
#> 622   0.4873069744      0
#> 623  -2.2549774615      0
#> 624  -0.9001367253      0
#> 625  -0.7101266574      0
#> 626   0.9080440443      0
#> 627   0.4996202695      0
#> 628   1.1618315426      0
#> 629  -1.3732285673      0
#> 630  -0.0032949517      0
#> 631   0.0345946011      0
#> 632  -2.4260552396      0
#> 633  -0.4487335436      0
#> 634   0.3911089232      0
#> 635   0.1523364732      0
#> 636   0.3564408779      0
#> 637  -0.7940557028      0
#> 638  -1.7693522626      0
#> 639  -0.1733315055      0
#> 640   0.2498659701      0
#> 641  -2.0438846424      0
#> 642  -1.7696906062      0
#> 643  -0.0411328397      0
#> 644   0.2942340187      0
#> 645   0.5157079754      0
#> 646  -0.7859471770      0
#> 647   0.3316329152      0
#> 648  -0.3357938182      0
#> 649   0.1018028891      0
#> 650   0.3201878181      0
#> 651  -1.7148680965      0
#> 652   0.8353281267      0
#> 653  -2.4054885921      0
#> 654   0.2395277053      0
#> 655  -0.2797035488      0
#> 656   0.2226540122      0
#> 657  -2.6985191419      0
#> 658  -0.7641718644      0
#> 659  -1.0682578405      0
#> 660   0.0865130966      0
#> 661  -1.0961414945      0
#> 662  -1.0586614542      0
#> 663  -0.5877881805      0
#> 664   0.8458420736      0
#> 665   1.1124357755      0
#> 666   0.8839727829      0
#> 667  -0.5470349151      0
#> 668  -1.2805167361      0
#> 669  -0.6815535753      0
#> 670  -0.2887000063      0
#> 671  -0.7406893795      0
#> 672   1.0330567280      0
#> 673   0.1729890030      0
#> 674  -0.0382800459      0
#> 675  -0.4012899853      0
#> 676  -1.1328127910      0
#> 677   0.5611911336      0
#> 678  -1.2021193023      0
#> 679  -0.9497621765      0
#> 680   1.0808180246      0
#> 681   1.3346578862      0
#> 682  -0.4153443256      0
#> 683   0.0379902253      0
#> 684  -1.3478200316      0
#> 685  -0.5943112589      0
#> 686  -0.6158483566      0
#> 687   0.6638471277      0
#> 688  -0.1136063178      0
#> 689   0.9977753429      0
#> 690  -0.8643078053      0
#> 691  -0.7202054620      0
#> 692   0.9966407420      0
#> 693   0.1493160531      0
#> 694   1.1031379749      0
#> 695  -2.5247168022      0
#> 696  -0.1022942638      0
#> 697   0.3204919698      0
#> 698   1.8416521653      0
#> 699  -1.0481106217      0
#> 700   0.3031703194      0
#> 701   0.4612882775      0
#> 702  -0.3944383803      0
#> 703  -0.9537260312      0
#> 704  -0.5762857860      0
#> 705   0.8401607186      0
#> 706   0.7760941312      0
#> 707  -0.9654006196      0
#> 708  -0.0910112610      0
#> 709   0.0487193869      0
#> 710   2.1949559902      0
#> 711   1.3505042342      0
#> 712  -0.6287619058      0
#> 713  -0.0260687830      0
#> 714   0.4587192583      0
#> 715  -0.0183059225      0
#> 716  -0.4148787029      0
#> 717  -1.1242859777      0
#> 718  -0.3734538907      0
#> 719  -0.0944307621      0
#> 720  -0.3241660949      0
#> 721  -1.8955962730      0
#> 722  -0.6605561246      0
#> 723   1.3321759698      0
#> 724  -0.5362773205      0
#> 725   1.0481594016      0
#> 726  -0.6338425864      0
#> 727   0.7092068541      0
#> 728  -0.7796653256      0
#> 729   1.3421851452      0
#> 730   0.2496917460      0
#> 731  -0.1206437890      0
#> 732   1.4711541172      0
#> 733   0.1707074578      0
#> 734  -0.2263208939      0
#> 735  -0.1826816968      0
#> 736  -0.8240448175      0
#> 737  -1.5521416061      0
#> 738   0.4992510896      0
#> 739  -1.3995551382      0
#> 740   0.5195750290      0
#> 741  -0.3303543368      0
#> 742  -0.0548953116      0
#> 743  -1.0915993713      0
#> 744   1.5189117399      0
#> 745   0.5074221041      0
#> 746  -1.3291747602      0
#> 747  -0.3464631608      0
#> 748   1.0888037469      0
#> 749   0.8951719117      0
#> 750   1.3520105886      0
#> 751  -0.1450161920      0
#> 752   1.1419205212      0
#> 753  -0.6350317558      0
#> 754   1.6860810508      0
#> 755  -1.6242611740      0
#> 756   0.5636041944      0
#> 757  -0.2671082522      0
#> 758  -0.0978238856      0
#> 759  -0.2988404920      0
#> 760   0.6380483462      0
#> 761   0.5466941013      0
#> 762   0.9017997271      0
#> 763   0.7409350624      0
#> 764  -0.4065183132      0
#> 765  -0.5508528877      0
#> 766   0.0041958375      0
#> 767   1.7754744113      0
#> 768  -0.3534064677      0
#> 769   0.0500466192      0
#> 770  -0.1222203780      0
#> 771   0.2417781302      0
#> 772  -1.4274003219      0
#> 773   0.2042286304      0
#> 774  -1.6721641285      0
#> 775  -1.0370673010      0
#> 776   1.2866560651      0
#> 777   0.3042994472      0
#> 778   0.4438908969      0
#> 779  -0.6610002192      0
#> 780   0.8353422484      0
#> 781   1.1021214898      0
#> 782   1.2319712483      0
#> 783   0.5078837517      0
#> 784  -1.2227663441      0
#> 785  -0.2552286514      0
#> 786   0.8903538762      0
#> 787   0.9559203932      0
#> 788  -1.0891740221      0
#> 789   0.4677503291      0
#> 790   0.3182310882      0
#> 791   0.2235711975      0
#> 792  -1.8614160905      0
#> 793  -1.8628256210      0
#> 794  -1.4232459597      0
#> 795   0.7696190744      0
#> 796   1.1322917897      0
#> 797   1.0359311439      0
#> 798   0.3653864903      0
#> 799  -0.3089180589      0
#> 800   1.2433957542      0
#> 801   0.9788360071      0
#> 802  -0.2729663499      0
#> 803   0.9924179589      0
#> 804   0.0495736576      0
#> 805  -1.2962361988      0
#> 806   1.0798816032      0
#> 807  -2.5535206868      0
#> 808   1.0704009020      0
#> 809  -1.8143976801      0
#> 810  -1.0936970334      0
#> 811  -0.1676600665      0
#> 812   0.5868672586      0
#> 813   0.0825222340      0
#> 814  -0.4591391413      0
#> 815  -1.0272334016      0
#> 816  -0.2463796250      0
#> 817   1.2830875234      0
#> 818  -0.2948579811      0
#> 819   0.4183337768      0
#> 820   0.9853573582      0
#> 821  -1.3779276137      0
#> 822   0.0497900232      0
#> 823   0.1088036590      0
#> 824  -0.1456689864      0
#> 825  -0.1456597293      0
#> 826   0.5908567561      0
#> 827  -0.7109934067      0
#> 828   0.4791598561      0
#> 829   0.0923249364      0
#> 830  -0.3138316838      0
#> 831  -0.1163501010      0
#> 832  -1.6241652696      0
#> 833  -0.4676229589      0
#> 834  -1.2967073170      0
#> 835   0.0956605664      0
#> 836   0.4319812770      0
#> 837  -1.3213376167      0
#> 838   1.1534636891      0
#> 839  -1.4235484988      0
#> 840  -1.5027594212      0
#> 841   1.2677159645      0
#> 842  -1.4201348188      0
#> 843   0.1266606369      0
#> 844   1.1368081483      0
#> 845  -1.3117169637      0
#> 846   2.4258411896      0
#> 847   1.2192888787      0
#> 848  -0.8020041277      0
#> 849   0.4969411559      0
#> 850   0.3912148717      0
#> 851  -1.3036820923      0
#> 852   1.9275764233      0
#> 853  -0.8660094861      0
#> 854   1.2976966160      0
#> 855  -0.0359203272      0
#> 856   0.4347014126      0
#> 857   0.2793195996      0
#> 858  -1.1421498940      0
#> 859   0.5322175169      0
#> 860   0.0113285451      0
#> 861   1.2692943010      0
#> 862   2.4095723414      0
#> 863  -0.9459522841      0
#> 864  -0.7092281510      0
#> 865  -1.3277996315      0
#> 866   0.7060289408      0
#> 867  -0.0676427383      0
#> 868  -0.5517895879      0
#> 869  -1.0906350575      0
#> 870   2.0271039716      0
#> 871  -0.4367536425      0
#> 872   0.9501757685      0
#> 873   1.4661184615      0
#> 874   0.1320283495      0
#> 875  -1.0915230481      0
#> 876  -2.1192555622      0
#> 877  -0.7084672959      0
#> 878  -0.5675673472      0
#> 879  -1.9735385840      0
#> 880   1.1736684857      0
#> 881   0.7321209482      0
#> 882  -2.5249207893      0
#> 883  -0.2718593068      0
#> 884  -0.4271878573      0
#> 885  -0.5243502993      0
#> 886  -0.0521400385      0
#> 887   0.9612404378      0
#> 888   0.2840549528      0
#> 889  -1.3995419276      0
#> 890  -0.2679998658      0
#> 891  -0.8952879187      0
#> 892   1.0926734696      0
#> 893   0.7081370033      0
#> 894   0.7376417493      0
#> 895  -1.0955909404      0
#> 896   0.8708765410      0
#> 897  -0.6843625972      0
#> 898   0.0286317884      0
#> 899   0.4030437604      0
#> 900  -1.0030295069      0
#> 901   0.6404271739      0
#> 902  -0.3322097792      0
#> 903  -1.7435185866      0
#> 904   0.9572175632      0
#> 905  -2.4388530909      0
#> 906   1.4246859577      0
#> 907  -1.2145624163      0
#> 908   1.2563741376      0
#> 909  -0.7248754135      0
#> 910   0.1240680983      0
#> 911  -1.6276745548      0
#> 912   0.2076712690      0
#> 913   0.8157922059      0
#> 914   1.4466491475      0
#> 915   0.7830882981      0
#> 916  -1.3361188479      0
#> 917  -0.6499268947      0
#> 918   1.1531551108      0
#> 919   0.1381249354      0
#> 920   0.0971166894      0
#> 921   0.3865252317      1
#> 922   0.4049651100      1
#> 923   2.2631138474      1
#> 924   1.2369810765      1
#> 925   0.7067309506      1
#> 926   2.4148442350      1
#> 927   0.9007618513      1
#> 928   0.1392780342      1
#> 929  -0.0538053708      1
#> 930   0.1858265946      1
#> 931  -0.4789444946      1
#> 932   1.1508004914      1
#> 933   0.3667716013      1
#> 934   1.1866539884      1
#> 935   2.5834834519      1
#> 936  -0.7092550338      1
#> 937   0.3494297240      1
#> 938   1.7889184115      1
#> 939   0.3561739246      1
#> 940   0.8888384730      1
#> 941  -0.3191098099      1
#> 942   0.0855262534      1
#> 943   2.7534164130      1
#> 944   1.7397983551      1
#> 945  -0.1176036341      1
#> 946   0.3483178197      1
#> 947   2.1631221241      1
#> 948   1.8830883343      1
#> 949   0.3314759528      1
#> 950   0.2292684985      1
#> 951   0.2020016637      1
#> 952  -1.2209908748      1
#> 953  -0.7531280256      1
#> 954   0.6003798971      1
#> 955  -0.6203845486      1
#> 956  -0.7063828944      1
#> 957  -0.5005392017      1
#> 958   0.4505593906      1
#> 959   2.3145275335      1
#> 960   0.8835314835      1
#> 961  -0.8506129610      1
#> 962   0.0163161310      1
#> 963   0.6417983541      1
#> 964  -0.5383652692      1
#> 965   0.3817709007      1
#> 966   3.1979862225      1
#> 967   1.3304609818      1
#> 968   0.4045095393      1
#> 969   3.2841781451      1
#> 970   0.1046423554      1
#> 971   0.6998208326      1
#> 972  -0.2624828214      1
#> 973   0.9309238641      1
#> 974  -0.5690370248      1
#> 975   0.2372645131      1
#> 976   0.3663363148      1
#> 977   0.5646302020      1
#> 978   0.8040130738      1
#> 979   0.3991800279      1
#> 980   2.0829803306      1
#> 981   0.0153392256      1
#> 982   0.6347303598      1
#> 983   1.3275881859      1
#> 984   1.9750269334      1
#> 985   0.6350930343      1
#> 986  -1.6598146755      1
#> 987  -0.0806582965      1
#> 988   0.2476075712      1
#> 989   1.3160699824      1
#> 990  -0.1245613505      1
#> 991   1.4480570699      1
#> 992  -0.0665867617      1
#> 993   0.4344841982      1
#> 994   1.6124698661      1
#> 995   0.3353191036      1
#> 996   1.1797349044      1
#> 997   0.8517012909      1
#> 998   0.8095843676      1
#> 999  -0.2541708231      1
#> 1000  1.3311362327      1