Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.428149845      0
#> 2     1.461536017      0
#> 3     0.657797533      0
#> 4     0.219326570      0
#> 5     0.559144383      0
#> 6    -0.884360159      0
#> 7     1.947293367      0
#> 8    -0.171551418      0
#> 9     1.220967241      0
#> 10   -0.575770618      0
#> 11   -0.231979402      0
#> 12    1.339942080      0
#> 13   -0.134032522      0
#> 14   -1.456943073      0
#> 15   -0.263871507      0
#> 16    0.259720563      0
#> 17    0.604675500      0
#> 18   -0.467653654      0
#> 19   -0.978172089      0
#> 20   -0.014913290      0
#> 21    0.311068787      0
#> 22   -1.712760150      0
#> 23    1.521897647      0
#> 24    1.300437910      0
#> 25   -1.357119018      0
#> 26   -0.392679426      0
#> 27    0.974876427      0
#> 28   -1.181085085      0
#> 29    0.345811542      0
#> 30    0.644149500      0
#> 31   -0.213269146      0
#> 32    0.337941841      0
#> 33    1.082318602      0
#> 34    0.969034912      0
#> 35    0.712014575      0
#> 36   -0.727232032      0
#> 37   -1.352884822      0
#> 38   -0.050249536      0
#> 39    0.870112264      0
#> 40   -0.121339716      0
#> 41   -1.037296112      0
#> 42    0.792952571      0
#> 43   -1.458276630      0
#> 44   -0.382452628      0
#> 45   -1.010055552      0
#> 46    0.166668236      0
#> 47   -0.873112769      0
#> 48   -1.312791477      0
#> 49    1.203308479      0
#> 50   -0.617862511      0
#> 51    0.392256885      0
#> 52    0.217841248      0
#> 53   -0.673022150      0
#> 54    0.925821386      0
#> 55   -0.559832058      0
#> 56   -0.794924360      0
#> 57    1.017213364      0
#> 58    0.553464145      0
#> 59    0.941252011      0
#> 60   -0.465981819      0
#> 61   -0.391912397      0
#> 62    0.260828115      0
#> 63   -0.146640537      0
#> 64    0.333173458      0
#> 65   -1.927258329      0
#> 66    0.257956740      0
#> 67    0.732729630      0
#> 68    1.044175205      0
#> 69    2.365251444      0
#> 70    0.283788496      0
#> 71   -0.717541312      0
#> 72    0.781491855      0
#> 73    0.597682683      0
#> 74   -1.743167129      0
#> 75   -0.316423600      0
#> 76   -0.897008779      0
#> 77   -0.984889953      0
#> 78   -2.142063407      0
#> 79   -0.752433141      0
#> 80    0.546082899      0
#> 81   -1.250825321      0
#> 82   -0.929649181      0
#> 83   -0.807872221      0
#> 84   -0.710511994      0
#> 85    0.344029931      0
#> 86   -0.748338159      0
#> 87   -0.581603314      0
#> 88   -0.856751596      0
#> 89   -2.158500321      0
#> 90    0.140644922      0
#> 91    1.244620268      0
#> 92   -0.757932934      0
#> 93   -2.029962625      0
#> 94    0.360356859      0
#> 95    0.060206196      0
#> 96   -0.584022922      0
#> 97   -2.489595632      0
#> 98    0.512369317      0
#> 99   -1.133194689      0
#> 100  -0.626228278      0
#> 101  -1.635172538      0
#> 102  -0.097874693      0
#> 103   0.203132496      0
#> 104  -0.194573784      0
#> 105  -0.307832578      0
#> 106  -0.636903473      0
#> 107  -2.146022066      0
#> 108   0.539897137      0
#> 109   0.123572199      0
#> 110   0.252711585      0
#> 111   0.127071484      0
#> 112  -0.029605486      0
#> 113  -0.284502776      0
#> 114  -0.267211547      0
#> 115  -0.578761304      0
#> 116   0.269073714      0
#> 117   0.625977860      0
#> 118  -0.435011574      0
#> 119   1.291025781      0
#> 120  -1.216906746      0
#> 121   0.483619257      0
#> 122  -0.364924743      0
#> 123  -2.303807132      0
#> 124  -0.929477549      0
#> 125   0.839230763      0
#> 126  -1.259491955      0
#> 127  -1.688918521      0
#> 128   0.537427683      0
#> 129   0.275575177      0
#> 130  -0.771188709      0
#> 131   0.213128438      0
#> 132  -0.091196742      0
#> 133  -1.664051553      0
#> 134   0.987603398      0
#> 135   0.023373124      0
#> 136  -0.062137406      0
#> 137  -0.849707057      0
#> 138   0.569760740      0
#> 139  -0.039266959      0
#> 140   0.630120292      0
#> 141   0.732169015      0
#> 142  -1.105140980      0
#> 143  -0.540295038      0
#> 144  -0.402577691      0
#> 145  -0.129228572      0
#> 146   1.857160265      0
#> 147   1.132479237      0
#> 148   2.605204768      0
#> 149   0.249905647      0
#> 150   0.232248893      0
#> 151   0.690823615      0
#> 152   1.754320626      0
#> 153  -0.643340153      0
#> 154   1.623422892      0
#> 155  -0.660050427      0
#> 156   1.633173574      0
#> 157   2.264727163      0
#> 158   0.156364305      0
#> 159  -0.175458921      0
#> 160  -0.475740612      0
#> 161   0.412743058      0
#> 162  -0.234798408      0
#> 163  -0.125556498      0
#> 164   0.497094198      0
#> 165  -0.542431947      0
#> 166  -0.116042237      0
#> 167   0.585763446      0
#> 168  -0.385054439      0
#> 169  -0.593223582      0
#> 170   0.230780766      0
#> 171   1.498543892      0
#> 172  -0.627898909      0
#> 173   2.022761301      0
#> 174   0.160980091      0
#> 175  -0.177516758      0
#> 176  -0.318906247      0
#> 177  -0.333456993      0
#> 178   1.104315873      0
#> 179   1.511390224      0
#> 180   0.849707782      0
#> 181  -0.284721212      0
#> 182   0.926064309      0
#> 183  -0.692404224      0
#> 184   2.071814588      0
#> 185  -0.245351175      0
#> 186   1.249889783      0
#> 187  -0.410222482      0
#> 188  -0.665919918      0
#> 189   0.176092940      0
#> 190   0.176944545      0
#> 191  -0.044382302      0
#> 192  -0.001924542      0
#> 193   0.231576991      0
#> 194   1.100437238      0
#> 195  -1.147581618      0
#> 196  -0.271605406      0
#> 197  -0.516302210      0
#> 198   1.585798080      0
#> 199  -0.624053951      0
#> 200   0.926942083      0
#> 201   0.437202917      0
#> 202  -0.377615108      0
#> 203   1.172913766      0
#> 204   1.925074432      0
#> 205   0.817297247      0
#> 206  -0.233865964      0
#> 207   0.523964401      0
#> 208  -0.289042351      0
#> 209   1.425928563      0
#> 210   0.157125271      0
#> 211  -0.953445686      0
#> 212   0.484411544      0
#> 213   0.551329937      0
#> 214   0.645662908      0
#> 215  -2.065560190      0
#> 216  -0.167061771      0
#> 217   0.004293205      0
#> 218   1.561556808      0
#> 219   0.180407139      0
#> 220  -0.947039228      0
#> 221   0.564523023      0
#> 222  -1.058091531      0
#> 223  -1.237528420      0
#> 224   1.222534717      0
#> 225  -1.923220355      0
#> 226  -0.884415334      0
#> 227   0.106716024      0
#> 228   0.344544540      0
#> 229  -0.868112484      0
#> 230  -0.537565326      0
#> 231   1.556570153      0
#> 232  -1.133169019      0
#> 233  -0.213783145      0
#> 234  -0.614530822      0
#> 235   0.176236656      0
#> 236  -1.557830554      0
#> 237  -0.524281600      0
#> 238   0.765341857      0
#> 239  -1.891820966      0
#> 240   1.043980790      0
#> 241  -0.159034494      0
#> 242  -0.799249319      0
#> 243  -0.673879584      0
#> 244  -0.681564883      0
#> 245  -0.456560489      0
#> 246   0.676492310      0
#> 247   0.897637072      0
#> 248   0.115902892      0
#> 249  -0.070483651      0
#> 250  -1.468253137      0
#> 251  -0.224459228      0
#> 252  -0.830976158      0
#> 253  -1.448328386      0
#> 254  -0.660057514      0
#> 255  -0.491109793      0
#> 256   0.642555737      0
#> 257  -0.019508231      0
#> 258  -0.335907935      0
#> 259  -1.628865377      0
#> 260   1.875303570      0
#> 261   0.127751896      0
#> 262  -0.142128795      0
#> 263   1.030506097      0
#> 264  -0.454930720      0
#> 265   0.177492718      0
#> 266  -0.454874099      0
#> 267  -0.472597746      0
#> 268   0.292834640      0
#> 269  -0.476918626      0
#> 270   0.262189996      0
#> 271  -1.547088463      0
#> 272  -0.967588381      0
#> 273   1.383498114      0
#> 274  -0.029445476      0
#> 275   1.354848087      0
#> 276  -0.542857337      0
#> 277  -0.840722825      0
#> 278   0.233970490      0
#> 279   1.311155237      0
#> 280   0.559633774      0
#> 281   0.159722386      0
#> 282  -0.406024126      0
#> 283  -0.621903269      0
#> 284  -0.013933215      0
#> 285  -1.901336639      0
#> 286  -0.395576578      0
#> 287   0.329921424      0
#> 288  -0.321576099      0
#> 289   0.793269348      0
#> 290  -0.898545900      0
#> 291  -2.844100947      0
#> 292   0.453081997      0
#> 293   0.508574978      0
#> 294   0.859200222      0
#> 295  -0.311416156      0
#> 296  -0.107921898      0
#> 297   0.086269880      0
#> 298  -0.428400923      0
#> 299   0.607584101      0
#> 300  -0.774627707      0
#> 301   0.278038961      0
#> 302  -0.144754696      0
#> 303   0.068517429      0
#> 304   0.588014847      0
#> 305  -0.346748228      0
#> 306   0.325605645      0
#> 307   1.091586574      0
#> 308  -0.311054429      0
#> 309  -1.659235480      0
#> 310  -0.132476860      0
#> 311  -0.295428405      0
#> 312   0.425213406      0
#> 313  -0.419226436      0
#> 314   1.223205438      0
#> 315   1.204510539      0
#> 316   0.449740909      0
#> 317  -1.249703285      0
#> 318   1.055161961      0
#> 319  -0.580167495      0
#> 320   0.348829154      0
#> 321  -0.579493758      0
#> 322   0.572876843      0
#> 323   0.558277121      0
#> 324  -0.802293668      0
#> 325   2.851983306      0
#> 326  -0.053460992      0
#> 327   0.066024650      0
#> 328  -0.706122423      0
#> 329  -0.087071019      0
#> 330   0.347033772      0
#> 331  -1.639686720      0
#> 332   0.536420108      0
#> 333  -1.410886460      0
#> 334  -1.052279757      0
#> 335   0.210203501      0
#> 336  -0.227686571      0
#> 337   0.978170733      0
#> 338  -0.226711321      0
#> 339   1.233502504      0
#> 340  -0.041674061      0
#> 341   0.540653012      0
#> 342   0.384614479      0
#> 343   0.891300808      0
#> 344   0.884199322      0
#> 345   0.303694542      0
#> 346  -0.442422808      0
#> 347   0.236333607      0
#> 348   1.166186063      0
#> 349  -0.098436384      0
#> 350   1.906752007      0
#> 351  -0.939183538      0
#> 352   0.814871083      0
#> 353  -0.647724571      0
#> 354   1.239111549      0
#> 355   1.783407024      0
#> 356   0.159732992      0
#> 357   0.570222200      0
#> 358  -0.613035176      0
#> 359  -0.239656471      0
#> 360  -1.481178682      0
#> 361  -0.128615497      0
#> 362   0.117159625      0
#> 363   0.315279221      0
#> 364  -0.260896119      0
#> 365   0.630113777      0
#> 366   0.283093594      0
#> 367  -0.969307904      0
#> 368   1.061207831      0
#> 369   1.066166850      0
#> 370   0.120377936      0
#> 371   0.007117872      0
#> 372   0.975306879      0
#> 373   0.147962754      0
#> 374   0.710092673      0
#> 375   0.070775574      0
#> 376  -1.535179144      0
#> 377  -0.107200753      0
#> 378   0.068719181      0
#> 379  -0.230989376      0
#> 380   0.019628249      0
#> 381   1.832616676      0
#> 382   1.031153461      0
#> 383  -0.904782100      0
#> 384  -0.712712282      0
#> 385   0.229293446      0
#> 386  -0.485355986      0
#> 387  -0.548519667      0
#> 388  -1.330995455      0
#> 389   0.049539208      0
#> 390  -0.575131227      0
#> 391  -1.350641166      0
#> 392   1.970572901      0
#> 393   1.734367841      0
#> 394  -0.494101275      0
#> 395   0.861174258      0
#> 396   0.427310221      0
#> 397   0.273804961      0
#> 398  -0.066277089      0
#> 399  -0.739834937      0
#> 400  -1.198266164      0
#> 401  -0.737957008      0
#> 402   1.422791400      0
#> 403   1.520095696      0
#> 404  -0.342816476      0
#> 405  -0.890296436      0
#> 406  -0.007474492      0
#> 407   0.553486797      0
#> 408  -0.728769156      0
#> 409   0.164602624      0
#> 410  -0.131053417      0
#> 411   0.269719784      0
#> 412   0.282355721      0
#> 413   2.501098847      0
#> 414   1.027792417      0
#> 415  -0.349112163      0
#> 416   0.167020915      0
#> 417   1.310629746      0
#> 418   0.164256655      0
#> 419  -0.950339627      0
#> 420   1.362586138      0
#> 421   0.935580616      0
#> 422   1.003560892      0
#> 423  -0.608007751      0
#> 424   0.696440859      0
#> 425  -1.043797396      0
#> 426   1.237218328      0
#> 427   0.008418138      0
#> 428   0.364247774      0
#> 429   1.079267404      0
#> 430  -1.864450807      0
#> 431  -0.070248455      0
#> 432  -2.951149153      0
#> 433  -1.144238255      0
#> 434  -0.432523127      0
#> 435  -1.634979412      0
#> 436  -0.120779399      0
#> 437   0.037121212      0
#> 438   0.595444689      0
#> 439   0.497510545      0
#> 440  -0.499419391      0
#> 441   0.702928359      0
#> 442  -2.819536312      0
#> 443   0.203260062      0
#> 444  -1.809640938      0
#> 445   3.185898120      0
#> 446  -0.384122044      0
#> 447  -0.177882532      0
#> 448   1.738747624      0
#> 449   0.280526610      0
#> 450  -0.750943247      0
#> 451   0.633439195      0
#> 452   0.833001266      0
#> 453   1.327149617      0
#> 454  -0.056313273      0
#> 455  -1.861754884      0
#> 456   0.068349127      0
#> 457   1.589404273      0
#> 458  -1.049440378      0
#> 459   0.266361351      0
#> 460  -1.111931664      0
#> 461   0.970874662      0
#> 462   0.685316212      0
#> 463   1.085915391      0
#> 464  -0.406620774      0
#> 465  -0.070061338      0
#> 466  -1.056444182      0
#> 467   1.794313493      0
#> 468  -1.943894399      0
#> 469  -0.389410262      0
#> 470  -0.647940534      0
#> 471  -0.741856642      0
#> 472   1.507444764      0
#> 473   0.428146845      0
#> 474   1.538479315      0
#> 475  -0.325891913      0
#> 476   1.534929258      0
#> 477   2.064686084      0
#> 478  -0.496692506      0
#> 479  -0.567959473      0
#> 480   0.273103929      0
#> 481   0.070968016      0
#> 482  -2.913735950      0
#> 483  -0.244382948      0
#> 484  -0.741528249      0
#> 485  -1.472910386      0
#> 486  -1.380749234      0
#> 487  -0.687394130      0
#> 488   0.034647858      0
#> 489  -0.724079893      0
#> 490  -0.368239071      0
#> 491   1.885626283      0
#> 492  -0.027637206      0
#> 493   0.769473264      0
#> 494   0.067668997      0
#> 495  -0.936321988      0
#> 496   0.753890231      0
#> 497  -0.005691889      0
#> 498   0.343813519      0
#> 499   0.476342168      0
#> 500  -0.155014186      0
#> 501   0.264851044      0
#> 502  -0.731590956      0
#> 503  -0.539943602      0
#> 504  -0.569278759      0
#> 505   0.372868893      0
#> 506   0.574355456      0
#> 507   0.295810660      0
#> 508  -0.037145923      0
#> 509  -1.164687210      0
#> 510  -0.282477790      0
#> 511   0.127132111      0
#> 512   0.857523457      0
#> 513   0.747066471      0
#> 514  -0.626638879      0
#> 515  -0.576930993      0
#> 516  -0.162160786      0
#> 517   1.070803820      0
#> 518  -0.107643254      0
#> 519  -0.674748238      0
#> 520  -1.048271224      0
#> 521  -1.012946864      0
#> 522   0.916449651      0
#> 523   0.612066282      0
#> 524   0.078699701      0
#> 525   0.868368638      0
#> 526  -0.102734490      0
#> 527   1.286498826      0
#> 528  -0.151693614      0
#> 529  -0.819241339      0
#> 530   0.182687352      0
#> 531   1.489613928      0
#> 532   0.823947669      0
#> 533   1.110906021      0
#> 534  -0.092867079      0
#> 535  -1.464081989      0
#> 536  -0.258587106      0
#> 537  -0.642952143      0
#> 538   0.185225367      0
#> 539  -0.016310180      0
#> 540   0.183571026      0
#> 541   0.980206135      0
#> 542  -0.422457745      0
#> 543  -0.490440948      0
#> 544  -0.287171516      0
#> 545  -0.854816710      0
#> 546   0.075865961      0
#> 547   2.204626290      0
#> 548   1.332945029      0
#> 549  -0.411664740      0
#> 550   0.408401139      0
#> 551  -0.836573514      0
#> 552   1.638500402      0
#> 553  -0.112968231      0
#> 554   0.741903153      0
#> 555   1.270652317      0
#> 556  -1.561172827      0
#> 557  -0.199112486      0
#> 558   0.262165999      0
#> 559  -0.893433812      0
#> 560   1.156924134      0
#> 561  -1.048468979      0
#> 562   0.370904630      0
#> 563   0.048452831      0
#> 564   0.639435161      0
#> 565   0.233867348      0
#> 566   1.062507993      0
#> 567  -1.025245332      0
#> 568  -0.236232316      0
#> 569   0.427744195      0
#> 570   2.950896484      0
#> 571   0.273357435      0
#> 572  -2.013220210      0
#> 573   0.463413540      0
#> 574  -0.794271689      0
#> 575  -0.124828843      0
#> 576   0.604914834      0
#> 577   0.312518639      0
#> 578  -0.251305371      0
#> 579   0.402812197      0
#> 580  -0.213593904      0
#> 581  -0.473449687      0
#> 582  -0.188468808      0
#> 583   1.454292881      0
#> 584  -0.665749246      0
#> 585   0.461110318      0
#> 586   0.140938643      0
#> 587   0.863692101      0
#> 588  -1.509023358      0
#> 589   1.145560525      0
#> 590   0.375605811      0
#> 591  -1.659807366      0
#> 592   0.900443835      0
#> 593   0.318362871      0
#> 594   0.367203333      0
#> 595  -0.631435931      0
#> 596   0.389111939      0
#> 597  -0.022612679      0
#> 598  -1.426590302      0
#> 599  -1.060073185      0
#> 600  -0.648915057      0
#> 601  -0.699746631      0
#> 602   0.179048500      0
#> 603   1.322417790      0
#> 604   0.191746586      0
#> 605   2.349736069      0
#> 606  -1.643442094      0
#> 607   2.370603380      0
#> 608  -0.632728232      0
#> 609  -0.571637158      0
#> 610   2.844484594      0
#> 611   0.118893421      0
#> 612   1.795455163      0
#> 613   0.154708914      0
#> 614   0.043503981      0
#> 615   0.253320980      0
#> 616  -0.828889282      0
#> 617   0.580326905      0
#> 618   0.286679313      0
#> 619   0.106760487      0
#> 620  -0.356908833      0
#> 621   1.141710598      0
#> 622  -1.775190993      0
#> 623  -1.160836982      0
#> 624  -1.102327206      0
#> 625  -0.869380171      0
#> 626  -0.707730722      0
#> 627   2.716900643      0
#> 628   0.810343993      0
#> 629   0.021375645      0
#> 630  -2.310886251      0
#> 631  -1.786503898      0
#> 632  -0.584553721      0
#> 633   0.613754902      0
#> 634  -0.461477408      0
#> 635  -0.383831468      0
#> 636   1.404405835      0
#> 637   1.220695039      0
#> 638   2.028488092      0
#> 639  -1.393624232      0
#> 640   0.101613845      0
#> 641   0.354138788      0
#> 642  -0.209999580      0
#> 643   0.440448728      0
#> 644  -0.283969950      0
#> 645  -0.103018723      0
#> 646   0.701623684      0
#> 647  -0.545473779      0
#> 648  -1.661946694      0
#> 649   0.759802210      0
#> 650   0.542511773      0
#> 651   1.400916266      0
#> 652   0.939153784      0
#> 653   1.625535561      0
#> 654   1.712834573      0
#> 655   1.250536616      0
#> 656  -1.189809495      0
#> 657   0.080081588      0
#> 658   0.903707282      0
#> 659  -0.203209758      0
#> 660  -0.781879973      0
#> 661  -0.512168866      0
#> 662  -0.181209083      0
#> 663  -0.684118835      0
#> 664  -0.091187153      0
#> 665   1.234321121      0
#> 666   1.430162280      0
#> 667  -0.803413211      0
#> 668   1.674372246      0
#> 669  -0.830560491      0
#> 670  -0.551074621      0
#> 671   0.078199818      0
#> 672  -0.619170704      0
#> 673  -0.107830927      0
#> 674  -2.370183608      0
#> 675  -0.595676569      0
#> 676   0.733765620      0
#> 677   0.057876069      0
#> 678  -1.179978134      0
#> 679  -1.830589677      0
#> 680  -2.924854237      0
#> 681  -0.470497729      0
#> 682   0.302192934      0
#> 683  -0.692518307      0
#> 684   0.487302524      0
#> 685  -1.126151456      0
#> 686  -1.559154824      0
#> 687  -0.484211171      0
#> 688   0.027414296      0
#> 689  -1.920797647      0
#> 690  -0.172284252      0
#> 691   0.796515760      0
#> 692  -1.241718246      0
#> 693  -0.346803544      0
#> 694  -0.801567490      0
#> 695  -1.047162833      0
#> 696   0.895950233      0
#> 697   0.922622595      0
#> 698  -0.202544131      0
#> 699  -0.330711317      0
#> 700   0.787937758      0
#> 701   0.495724018      0
#> 702   0.545325844      0
#> 703  -0.269029631      0
#> 704  -1.570342261      0
#> 705   2.392794383      0
#> 706   0.382388812      0
#> 707   2.550291950      0
#> 708   0.880826340      0
#> 709   0.415316442      0
#> 710   0.208741604      0
#> 711   1.334773576      0
#> 712   0.821861587      0
#> 713  -1.042303774      0
#> 714   1.591390501      0
#> 715  -0.374061086      0
#> 716   0.056975065      0
#> 717  -0.477860211      0
#> 718  -1.511350721      0
#> 719   0.438515581      0
#> 720   2.901803005      0
#> 721   0.134177816      0
#> 722   0.880831770      0
#> 723  -0.572631039      0
#> 724   1.218869329      0
#> 725  -1.815104512      0
#> 726  -2.124206093      0
#> 727  -0.220007520      0
#> 728   0.065782908      0
#> 729   1.184280031      0
#> 730   0.987764725      0
#> 731   1.477998799      0
#> 732  -1.659659442      0
#> 733  -0.456401884      0
#> 734  -0.094456205      0
#> 735  -2.169999494      0
#> 736  -0.882926234      0
#> 737  -0.808529729      0
#> 738   0.443278881      0
#> 739  -0.653362104      0
#> 740  -1.347931277      0
#> 741  -0.079823359      0
#> 742  -0.475115203      0
#> 743  -0.812072865      0
#> 744  -0.714638993      0
#> 745   1.951717100      0
#> 746   0.306753237      0
#> 747   0.231536990      0
#> 748   1.087234529      0
#> 749   0.965618945      0
#> 750   1.050863238      0
#> 751  -0.760157244      0
#> 752  -0.955147743      0
#> 753  -1.871210764      0
#> 754  -0.733312630      0
#> 755   0.997869941      0
#> 756   1.156398791      0
#> 757  -0.562109668      0
#> 758  -0.614667947      0
#> 759  -0.213719967      0
#> 760  -1.828714251      0
#> 761  -1.157390843      0
#> 762  -0.498961229      0
#> 763  -1.970994192      0
#> 764  -0.214335672      0
#> 765   1.514899338      0
#> 766  -0.795795428      0
#> 767  -0.240946862      0
#> 768  -0.218990884      0
#> 769  -1.192779052      0
#> 770   0.111276063      0
#> 771  -2.029524359      0
#> 772  -0.795857309      0
#> 773  -0.503564414      0
#> 774   0.546601713      0
#> 775  -0.153137749      0
#> 776   1.428968806      0
#> 777  -1.604431291      0
#> 778   1.459840091      0
#> 779  -0.283021577      0
#> 780  -0.773783363      0
#> 781   0.075692445      0
#> 782  -0.916099062      0
#> 783  -1.110940841      0
#> 784  -1.516555600      0
#> 785  -1.971859825      0
#> 786   0.396407828      0
#> 787  -0.452898954      0
#> 788  -0.389375089      0
#> 789   0.544264876      0
#> 790  -0.997962149      0
#> 791   1.458090914      0
#> 792  -0.039330379      0
#> 793   0.266309703      0
#> 794   0.340819906      0
#> 795   0.061294605      0
#> 796  -1.126702848      0
#> 797  -0.847687469      0
#> 798   0.142706117      0
#> 799  -0.488117007      0
#> 800   1.261690955      0
#> 801   2.487762218      0
#> 802  -0.009669133      0
#> 803   0.068924465      0
#> 804   0.602022204      0
#> 805  -0.028771247      0
#> 806   0.211388159      0
#> 807  -2.144492757      0
#> 808  -0.232152791      0
#> 809  -1.145837900      0
#> 810  -0.964508657      0
#> 811   1.255546647      0
#> 812   0.227616331      0
#> 813   1.424195959      0
#> 814  -0.120667191      0
#> 815  -1.784729453      0
#> 816  -0.382775855      0
#> 817   1.229285148      0
#> 818   0.272458117      0
#> 819  -0.114396711      0
#> 820  -0.153104250      0
#> 821   0.093624792      0
#> 822   2.683350121      0
#> 823   0.488947098      0
#> 824   0.726943096      0
#> 825  -0.639002005      0
#> 826   0.026927236      0
#> 827  -0.521378532      0
#> 828   0.773482201      0
#> 829   0.230006924      0
#> 830  -0.737301810      0
#> 831  -0.101612248      0
#> 832   1.211180845      0
#> 833  -0.570950922      0
#> 834   2.468065109      0
#> 835  -0.022061416      0
#> 836   0.283753048      0
#> 837   0.655884612      0
#> 838   0.757448444      0
#> 839  -0.843661893      0
#> 840  -0.945407103      0
#> 841  -1.359810955      0
#> 842  -1.046423480      0
#> 843   0.135035286      0
#> 844  -0.846498535      0
#> 845  -0.227185182      0
#> 846   1.201366991      0
#> 847   0.238223156      0
#> 848   0.727187525      0
#> 849   1.198997994      0
#> 850   0.619672392      0
#> 851  -1.413732243      0
#> 852   0.408709321      0
#> 853  -0.210026244      0
#> 854  -0.453047818      0
#> 855   2.647943074      0
#> 856  -0.103587922      0
#> 857  -0.473353747      0
#> 858  -0.183871338      0
#> 859  -0.566748266      0
#> 860   1.708373831      0
#> 861  -1.355185169      0
#> 862   1.397728595      0
#> 863  -0.089222823      0
#> 864   1.111152741      0
#> 865   0.567829733      0
#> 866   1.852113043      0
#> 867  -0.779897888      0
#> 868  -0.404700577      0
#> 869  -0.273327293      0
#> 870  -1.544096128      0
#> 871  -0.164874065      0
#> 872   0.982599307      0
#> 873  -0.234713084      0
#> 874  -0.416563185      0
#> 875   0.298327255      0
#> 876  -1.299458510      0
#> 877   0.498905818      0
#> 878  -0.729495710      0
#> 879   0.021518273      0
#> 880   0.228169411      0
#> 881   0.269890513      0
#> 882   0.946049782      0
#> 883   0.401712473      0
#> 884  -0.897175698      0
#> 885  -0.345679663      0
#> 886   0.102094010      0
#> 887  -0.992982793      0
#> 888   0.175617674      0
#> 889  -0.873681102      0
#> 890   0.290375025      0
#> 891  -1.532079765      0
#> 892   1.536320704      0
#> 893  -1.607071006      0
#> 894  -1.418418505      0
#> 895   0.008704474      0
#> 896  -0.833445543      0
#> 897  -2.157746349      0
#> 898   0.698497184      0
#> 899   1.955729137      0
#> 900  -0.057401412      0
#> 901   2.051792260      0
#> 902  -1.350716755      0
#> 903  -0.653089300      0
#> 904   0.026080893      0
#> 905  -1.555615486      0
#> 906   0.542502478      0
#> 907   0.261001045      0
#> 908   1.824410711      1
#> 909   2.015132894      1
#> 910   1.360649680      1
#> 911  -0.168361378      1
#> 912   2.206924749      1
#> 913   0.299492541      1
#> 914   2.209529507      1
#> 915   1.265976645      1
#> 916   1.285589861      1
#> 917  -0.473777442      1
#> 918   2.048274926      1
#> 919   1.306317852      1
#> 920   1.138688309      1
#> 921  -0.117892597      1
#> 922   0.598549720      1
#> 923   3.354232470      1
#> 924   1.602511569      1
#> 925   1.061855854      1
#> 926   2.210205216      1
#> 927  -0.635843207      1
#> 928   1.527201698      1
#> 929   1.250529329      1
#> 930   0.724311440      1
#> 931   2.292998025      1
#> 932   0.278539828      1
#> 933   1.098807561      1
#> 934   0.216666942      1
#> 935   1.135084356      1
#> 936   2.045548005      1
#> 937   0.996077636      1
#> 938  -2.021447107      1
#> 939   1.069892987      1
#> 940   0.654262377      1
#> 941   1.224452468      1
#> 942  -0.200001127      1
#> 943   1.113136535      1
#> 944  -1.923407002      1
#> 945   0.260763155      1
#> 946   3.170618148      1
#> 947  -1.653071154      1
#> 948   1.599308339      1
#> 949   0.716756919      1
#> 950   0.768391824      1
#> 951   0.107943039      1
#> 952   0.397202942      1
#> 953   0.540489844      1
#> 954   1.459690086      1
#> 955   0.489284578      1
#> 956   0.909649518      1
#> 957   1.018169693      1
#> 958   1.298069049      1
#> 959  -0.366605693      1
#> 960   0.917867928      1
#> 961   1.269474710      1
#> 962   0.727531680      1
#> 963  -2.423254394      1
#> 964   1.202178135      1
#> 965   1.821055448      1
#> 966  -0.812976399      1
#> 967   1.258873944      1
#> 968   0.963038931      1
#> 969  -0.226456412      1
#> 970   2.618461842      1
#> 971   0.490086920      1
#> 972   1.511724534      1
#> 973   0.800756461      1
#> 974  -0.033370858      1
#> 975   1.830143679      1
#> 976   0.493414881      1
#> 977   0.638820999      1
#> 978  -0.690537865      1
#> 979   1.930250186      1
#> 980  -0.545826341      1
#> 981   1.169047942      1
#> 982  -1.099438851      1
#> 983  -1.381530814      1
#> 984   0.327848764      1
#> 985  -0.899008029      1
#> 986   1.956448017      1
#> 987   1.130362537      1
#> 988   1.820171124      1
#> 989   1.338827405      1
#> 990   0.973294651      1
#> 991   1.582541722      1
#> 992  -0.031464562      1
#> 993   1.630071177      1
#> 994  -0.566096908      1
#> 995   2.139177818      1
#> 996   1.347697976      1
#> 997   0.501548329      1
#> 998   0.603868690      1
#> 999   0.245131975      1
#> 1000 -0.077040975      1