Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.372603669      0
#> 2    -0.433765557      0
#> 3    -1.223230880      0
#> 4     0.458392970      0
#> 5     0.346965323      0
#> 6     2.666432451      0
#> 7     0.093901401      0
#> 8    -2.272951250      0
#> 9     1.839405963      0
#> 10    0.369232547      0
#> 11    0.565458502      0
#> 12   -1.795674469      0
#> 13    1.400047098      0
#> 14   -1.338488773      0
#> 15    0.038886768      0
#> 16   -1.171773639      0
#> 17    0.650857639      0
#> 18   -0.019649309      0
#> 19    0.232448136      0
#> 20   -0.868836687      0
#> 21    0.437166150      0
#> 22   -0.564693156      0
#> 23    0.890052214      0
#> 24    1.255813263      0
#> 25    1.354958596      0
#> 26    0.478814190      0
#> 27    1.928102263      0
#> 28    1.017119319      0
#> 29   -0.604045329      0
#> 30    0.105867299      0
#> 31   -2.240034333      0
#> 32   -1.213317663      0
#> 33   -0.539936927      0
#> 34    0.786315908      0
#> 35    0.934758501      0
#> 36    0.142407279      0
#> 37    0.082581653      0
#> 38    1.063961619      0
#> 39    1.433169372      0
#> 40   -0.239886035      0
#> 41    0.336704981      0
#> 42   -0.321379094      0
#> 43    1.056653918      0
#> 44   -1.171136855      0
#> 45   -0.682098251      0
#> 46    1.353504637      0
#> 47    2.385169106      0
#> 48    1.153044005      0
#> 49    0.192794528      0
#> 50    0.097586652      0
#> 51    0.758000640      0
#> 52   -0.626387570      0
#> 53   -1.427403285      0
#> 54   -0.361052148      0
#> 55    0.636559054      0
#> 56   -0.927468275      0
#> 57    0.073340543      0
#> 58   -0.386620998      0
#> 59   -0.492465919      0
#> 60   -0.589114814      0
#> 61   -1.567331917      0
#> 62   -0.259060970      0
#> 63    0.902112939      0
#> 64    0.162053171      0
#> 65    0.934684126      0
#> 66   -0.133010541      0
#> 67    0.482383411      0
#> 68   -1.472523967      0
#> 69    0.468409100      0
#> 70   -0.724259566      0
#> 71    1.154176506      0
#> 72   -0.984734747      0
#> 73   -1.594591396      0
#> 74    1.330193853      0
#> 75   -1.125686844      0
#> 76    0.855372631      0
#> 77    0.347155710      0
#> 78    2.219487507      0
#> 79   -0.900168468      0
#> 80    0.293574973      0
#> 81   -0.943728777      0
#> 82    0.435860912      0
#> 83   -0.928297783      0
#> 84    0.418515586      0
#> 85   -0.470168150      0
#> 86    0.122891950      0
#> 87    0.583967551      0
#> 88    0.950656770      0
#> 89    1.199629312      0
#> 90   -0.466196637      0
#> 91    1.037808470      0
#> 92   -0.245062491      0
#> 93    0.012153183      0
#> 94   -0.076208155      0
#> 95    0.614610940      0
#> 96   -0.160085660      0
#> 97   -1.064879091      0
#> 98    2.687088962      0
#> 99   -0.289686628      0
#> 100   1.445379686      0
#> 101  -0.380584268      0
#> 102   0.114335796      0
#> 103  -0.698600962      0
#> 104   0.505332352      0
#> 105   0.175575917      0
#> 106   0.926003363      0
#> 107  -1.056172779      0
#> 108   0.336318469      0
#> 109  -0.597275231      0
#> 110  -0.715910358      0
#> 111  -0.692935390      0
#> 112  -1.543026693      0
#> 113   0.509513974      0
#> 114   1.182164642      0
#> 115   1.937160636      0
#> 116  -0.873343386      0
#> 117  -0.715337414      0
#> 118   1.207379332      0
#> 119   0.141959604      0
#> 120  -0.551188918      0
#> 121  -1.020658492      0
#> 122   0.835101980      0
#> 123   0.086091657      0
#> 124   0.364976241      0
#> 125  -1.151970850      0
#> 126  -0.483500022      0
#> 127   0.551514604      0
#> 128  -1.348430706      0
#> 129  -0.055747179      0
#> 130  -0.406816154      0
#> 131   0.307793292      0
#> 132   0.330851118      0
#> 133  -0.289693546      0
#> 134   0.516360754      0
#> 135  -0.166221959      0
#> 136   0.360549639      0
#> 137   0.057633548      0
#> 138   1.985163404      0
#> 139  -0.823760685      0
#> 140   1.193327103      0
#> 141   0.107243246      0
#> 142   1.046855248      0
#> 143   0.106000902      0
#> 144   0.600774755      0
#> 145  -0.449985631      0
#> 146  -0.891306779      0
#> 147   1.269579105      0
#> 148  -0.957596521      0
#> 149   0.487999215      0
#> 150   0.370397662      0
#> 151  -0.295282307      0
#> 152   0.215550282      0
#> 153   0.219460547      0
#> 154   0.316343877      0
#> 155   0.672169453      0
#> 156  -1.457544666      0
#> 157  -0.323238031      0
#> 158   0.011975279      0
#> 159  -0.182401476      0
#> 160  -1.630303911      0
#> 161   0.605518056      0
#> 162   1.287694119      0
#> 163  -0.110593738      0
#> 164  -0.827226586      0
#> 165  -1.053629201      0
#> 166   1.577882582      0
#> 167  -1.780521812      0
#> 168  -0.683583775      0
#> 169  -0.546499327      0
#> 170   0.482517242      0
#> 171   0.423284958      0
#> 172   0.399477494      0
#> 173   1.457323848      0
#> 174   0.018244219      0
#> 175   0.457550745      0
#> 176  -1.358333119      0
#> 177   0.958478985      0
#> 178   0.373313932      0
#> 179   1.232895747      0
#> 180  -2.297059071      0
#> 181   0.394924541      0
#> 182  -0.352847070      0
#> 183   0.076526526      0
#> 184  -1.776136605      0
#> 185   0.906784531      0
#> 186  -0.530032532      0
#> 187   0.914725808      0
#> 188   0.895792973      0
#> 189  -0.405954990      0
#> 190   0.327861613      0
#> 191   0.354858808      0
#> 192   0.654461368      0
#> 193  -0.391199278      0
#> 194   0.295494771      0
#> 195   0.401467764      0
#> 196   0.436795368      0
#> 197   0.085057706      0
#> 198   1.132323235      0
#> 199   0.184651096      0
#> 200   0.178685158      0
#> 201   0.187795084      0
#> 202  -0.449235795      0
#> 203  -0.964806022      0
#> 204  -0.169848089      0
#> 205  -1.169384087      0
#> 206   0.106412857      0
#> 207  -0.278966011      0
#> 208  -0.851679515      0
#> 209   1.003991711      0
#> 210  -0.057411395      0
#> 211  -0.124050072      0
#> 212   1.288792367      0
#> 213   0.108901753      0
#> 214   0.626243722      0
#> 215  -2.430148138      0
#> 216  -0.497270673      0
#> 217  -1.354379358      0
#> 218   0.330740857      0
#> 219  -0.208519047      0
#> 220  -0.807803717      0
#> 221   0.464866636      0
#> 222   0.390874098      0
#> 223   0.126652016      0
#> 224   1.265020098      0
#> 225  -0.451531476      0
#> 226  -0.959385403      0
#> 227   0.125515428      0
#> 228  -0.227098335      0
#> 229   1.330392403      0
#> 230   0.058213721      0
#> 231  -0.064165187      0
#> 232   1.592971501      0
#> 233   0.585081292      0
#> 234   0.825831578      0
#> 235  -1.345805318      0
#> 236  -1.328773036      0
#> 237   1.903203101      0
#> 238   0.521799747      0
#> 239   0.754145013      0
#> 240  -1.253872915      0
#> 241  -0.850205514      0
#> 242  -0.937812390      0
#> 243   0.527487682      0
#> 244  -1.042411091      0
#> 245   1.669665267      0
#> 246  -0.945342215      0
#> 247  -1.517261516      0
#> 248   1.019820813      0
#> 249  -0.920338583      0
#> 250  -1.419798948      0
#> 251   0.372502704      0
#> 252  -0.149909785      0
#> 253   1.397510432      0
#> 254   2.278858397      0
#> 255   0.285522497      0
#> 256  -1.338234946      0
#> 257  -1.029474858      0
#> 258  -0.104024736      0
#> 259   0.788780289      0
#> 260  -0.393303989      0
#> 261  -0.025432431      0
#> 262  -0.540285658      0
#> 263   2.074727092      0
#> 264  -0.657328389      0
#> 265   0.078056386      0
#> 266  -0.201485953      0
#> 267   0.608440180      0
#> 268   0.381274619      0
#> 269  -0.334667456      0
#> 270   1.591639369      0
#> 271   1.583761404      0
#> 272  -0.456735719      0
#> 273   0.502988740      0
#> 274   1.462326356      0
#> 275  -1.509475929      0
#> 276  -0.321468397      0
#> 277  -0.820474662      0
#> 278   0.971783466      0
#> 279  -0.900410904      0
#> 280  -1.715761856      0
#> 281   2.245155573      0
#> 282  -0.497683059      0
#> 283   2.546392899      0
#> 284   0.276707516      0
#> 285   0.056051074      0
#> 286  -2.387650595      0
#> 287  -1.153608827      0
#> 288  -1.037244316      0
#> 289  -0.272590172      0
#> 290   1.515553067      0
#> 291   0.891874688      0
#> 292  -0.627660620      0
#> 293   1.246873658      0
#> 294  -1.139572797      0
#> 295  -0.265162638      0
#> 296  -0.469010908      0
#> 297  -0.575568381      0
#> 298   0.630606679      0
#> 299  -0.127201043      0
#> 300  -0.334932099      0
#> 301  -0.736227858      0
#> 302  -1.138945204      0
#> 303   0.562878365      0
#> 304  -0.307644983      0
#> 305  -0.348367982      0
#> 306   0.537340258      0
#> 307   1.936788986      0
#> 308   0.023086108      0
#> 309   1.504979553      0
#> 310   0.233734659      0
#> 311  -0.105517086      0
#> 312   0.870093842      0
#> 313   0.277115509      0
#> 314   0.086143575      0
#> 315   0.944949291      0
#> 316  -0.438646013      0
#> 317  -0.199287561      0
#> 318  -0.292761757      0
#> 319  -0.635408969      0
#> 320  -1.055788842      0
#> 321  -0.009695033      0
#> 322   2.004617601      0
#> 323  -1.823933759      0
#> 324   0.157190128      0
#> 325  -0.146563384      0
#> 326  -0.101159264      0
#> 327  -0.472847163      0
#> 328   0.683856549      0
#> 329  -0.653120526      0
#> 330   0.145008549      0
#> 331  -1.342020829      0
#> 332   0.812756164      0
#> 333   0.865357237      0
#> 334   1.414655780      0
#> 335   1.509142392      0
#> 336   2.035850963      0
#> 337  -0.370118160      0
#> 338  -1.057017831      0
#> 339  -0.235447457      0
#> 340  -1.927318837      0
#> 341  -0.379323010      0
#> 342  -0.147497360      0
#> 343   0.651644202      0
#> 344   0.313154356      0
#> 345   0.015185934      0
#> 346   0.847199833      0
#> 347   0.108112527      0
#> 348  -1.326630234      0
#> 349   0.825422141      0
#> 350  -0.472473716      0
#> 351   0.183525968      0
#> 352   0.495153775      0
#> 353  -0.324981362      0
#> 354  -0.881654737      0
#> 355   0.528315249      0
#> 356  -0.808633942      0
#> 357   0.509771331      0
#> 358   0.112349706      0
#> 359   0.194370757      0
#> 360  -0.964957321      0
#> 361   0.693561946      0
#> 362  -0.054756116      0
#> 363  -0.564079201      0
#> 364   0.032135234      0
#> 365   0.154210881      0
#> 366  -0.239964449      0
#> 367  -0.057296672      0
#> 368  -0.266263317      0
#> 369  -0.976176165      0
#> 370   0.933747024      0
#> 371  -0.235043459      0
#> 372  -0.407530699      0
#> 373  -1.848383795      0
#> 374   0.473785466      0
#> 375  -0.460932664      0
#> 376   0.713224125      0
#> 377  -0.349009270      0
#> 378   0.095979109      0
#> 379   0.702131568      0
#> 380   1.597543847      0
#> 381  -0.277640122      0
#> 382   0.037897659      0
#> 383   0.385861921      0
#> 384   0.580450828      0
#> 385   0.671898304      0
#> 386  -0.323073535      0
#> 387   1.465462083      0
#> 388  -0.960620232      0
#> 389  -0.431472658      0
#> 390   0.180822022      0
#> 391  -0.419604185      0
#> 392   0.851289728      0
#> 393  -0.586629646      0
#> 394   0.780938642      0
#> 395  -0.298771740      0
#> 396  -0.437076493      0
#> 397   0.326507644      0
#> 398  -0.385523709      0
#> 399   0.209582396      0
#> 400   0.321766995      0
#> 401  -1.722578758      0
#> 402  -0.309770573      0
#> 403  -1.328183825      0
#> 404  -0.160894282      0
#> 405   0.162392811      0
#> 406   0.185084508      0
#> 407   0.653020459      0
#> 408  -0.002447794      0
#> 409   1.017293485      0
#> 410  -0.240396329      0
#> 411  -0.609647600      0
#> 412   0.086251147      0
#> 413   1.772396016      0
#> 414   0.298973315      0
#> 415   1.232485362      0
#> 416  -1.698156539      0
#> 417   0.478027205      0
#> 418   0.765442201      0
#> 419   0.899838294      0
#> 420   1.551995698      0
#> 421   0.879896735      0
#> 422  -0.906210805      0
#> 423  -0.161364691      0
#> 424  -0.704137971      0
#> 425  -0.665436766      0
#> 426  -1.317977889      0
#> 427   0.218245265      0
#> 428  -1.664220149      0
#> 429  -0.087974627      0
#> 430  -0.554662265      0
#> 431   0.081991938      0
#> 432  -0.571449365      0
#> 433  -1.721840714      0
#> 434   1.709886776      0
#> 435   1.757164852      0
#> 436   2.286124067      0
#> 437  -1.127121885      0
#> 438   1.078280871      0
#> 439   0.741981112      0
#> 440  -0.513931739      0
#> 441   0.637706791      0
#> 442   1.182054489      0
#> 443   0.848268485      0
#> 444   0.070628717      0
#> 445  -0.860106215      0
#> 446   0.536859516      0
#> 447   1.858757317      0
#> 448   0.933418428      0
#> 449   0.653582992      0
#> 450   1.077878670      0
#> 451  -1.949226399      0
#> 452   0.174672676      0
#> 453   1.072879263      0
#> 454  -0.624272142      0
#> 455  -0.030489361      0
#> 456   0.422001366      0
#> 457   0.757289702      0
#> 458   0.128321750      0
#> 459  -0.161855087      0
#> 460  -0.416126320      0
#> 461   0.287228868      0
#> 462  -0.130988890      0
#> 463  -2.658488487      0
#> 464   0.128274880      0
#> 465  -1.833632359      0
#> 466  -0.423921355      0
#> 467  -0.642033336      0
#> 468   0.735546430      0
#> 469   0.679130426      0
#> 470   0.465974036      0
#> 471   0.689565301      0
#> 472   1.388849547      0
#> 473   1.002518835      0
#> 474   0.157568554      0
#> 475  -1.781852885      0
#> 476  -0.023394040      0
#> 477  -1.480530683      0
#> 478  -0.559175139      0
#> 479   1.521185154      0
#> 480  -0.221379617      0
#> 481  -0.446880124      0
#> 482   0.744338660      0
#> 483   0.317367453      0
#> 484  -0.606670715      0
#> 485   1.121192846      0
#> 486   0.687211951      0
#> 487   0.620347978      0
#> 488   0.913729412      0
#> 489  -1.024093320      0
#> 490   1.118413460      0
#> 491   0.619499556      0
#> 492   0.066745050      0
#> 493   0.853825466      0
#> 494   0.569212755      0
#> 495  -0.395602771      0
#> 496  -1.444718286      0
#> 497   0.469286816      0
#> 498  -0.727808363      0
#> 499   1.793354570      0
#> 500  -0.149660776      0
#> 501  -0.227520133      0
#> 502   0.389326777      0
#> 503  -0.913671991      0
#> 504   1.237493493      0
#> 505   2.333141291      0
#> 506  -1.600532042      0
#> 507  -1.225153435      0
#> 508  -0.579418941      0
#> 509  -0.113926872      0
#> 510   0.128242094      0
#> 511   0.050386803      0
#> 512  -1.697880338      0
#> 513  -1.294534885      0
#> 514   0.742875152      0
#> 515   0.393238440      0
#> 516  -0.942994718      0
#> 517   0.558249209      0
#> 518  -1.015853583      0
#> 519   0.378956610      0
#> 520  -0.624264845      0
#> 521   0.189748605      0
#> 522  -0.234341124      0
#> 523  -0.090105541      0
#> 524   0.118276312      0
#> 525  -0.668082160      0
#> 526  -0.708080765      0
#> 527   0.818217273      0
#> 528   0.243395292      0
#> 529   0.270778828      0
#> 530  -1.208633339      0
#> 531  -1.383295125      0
#> 532   0.452394977      0
#> 533  -0.355545685      0
#> 534   0.939265073      0
#> 535  -0.323113558      0
#> 536  -1.463138810      0
#> 537  -0.082230570      0
#> 538   0.478101319      0
#> 539  -1.776048994      0
#> 540  -0.842350284      0
#> 541  -0.235808664      0
#> 542   0.116191867      0
#> 543  -0.811963709      0
#> 544  -1.175747206      0
#> 545   1.190531656      0
#> 546  -0.895163125      0
#> 547   1.582073640      0
#> 548   0.634218496      0
#> 549  -0.686922326      0
#> 550   1.072798544      0
#> 551  -0.453011546      0
#> 552  -1.029982907      0
#> 553  -2.698631094      0
#> 554  -0.280480680      0
#> 555  -2.803387351      0
#> 556  -1.413644317      0
#> 557   0.615562251      0
#> 558   0.429049113      0
#> 559  -0.370437462      0
#> 560   0.429089019      0
#> 561   0.533260720      0
#> 562  -0.884624654      0
#> 563   0.606008266      0
#> 564   1.376571113      0
#> 565   0.825538711      0
#> 566  -1.032924607      0
#> 567   1.349798968      0
#> 568   1.804173712      0
#> 569   0.796462652      0
#> 570  -0.072003962      0
#> 571  -0.306821860      0
#> 572  -0.624834762      0
#> 573  -1.818415601      0
#> 574   0.733221844      0
#> 575   0.548423033      0
#> 576  -0.215684345      0
#> 577  -1.472131521      0
#> 578  -0.204291986      0
#> 579  -0.201588034      0
#> 580  -0.309869035      0
#> 581   0.194242270      0
#> 582   0.581033217      0
#> 583   0.629498991      0
#> 584   1.683080298      0
#> 585   0.267591187      0
#> 586   0.408791382      0
#> 587  -0.318659372      0
#> 588   1.700045409      0
#> 589   0.954783517      0
#> 590  -1.105124371      0
#> 591  -0.585054372      0
#> 592  -0.396164600      0
#> 593   0.388918569      0
#> 594   1.028980323      0
#> 595   1.249844632      0
#> 596  -1.836768545      0
#> 597  -1.042946604      0
#> 598  -0.073465325      0
#> 599   0.132158762      0
#> 600   0.187211319      0
#> 601  -0.311598226      0
#> 602  -0.002026066      0
#> 603   1.458019768      0
#> 604  -0.384693352      0
#> 605   0.051530034      0
#> 606   0.065470988      0
#> 607  -0.209047966      0
#> 608   1.271058100      0
#> 609  -0.325788913      0
#> 610  -0.993112296      0
#> 611   0.073368995      0
#> 612  -3.364887617      0
#> 613   0.122543745      0
#> 614  -0.141640171      0
#> 615  -0.753925963      0
#> 616  -1.129567736      0
#> 617  -0.659331605      0
#> 618  -0.321709427      0
#> 619  -0.978895348      0
#> 620  -0.429605829      0
#> 621   2.340108533      0
#> 622   0.605488800      0
#> 623  -0.330441303      0
#> 624  -0.054499023      0
#> 625   1.417216931      0
#> 626   0.161662914      0
#> 627   0.144963602      0
#> 628  -1.434294840      0
#> 629  -0.922607749      0
#> 630  -0.552895698      0
#> 631   0.221658276      0
#> 632  -0.391272945      0
#> 633  -1.884276230      0
#> 634  -0.434389821      0
#> 635   0.383149460      0
#> 636  -0.908963194      0
#> 637  -0.477737210      0
#> 638  -1.289845578      0
#> 639  -0.323478704      0
#> 640  -0.833617782      0
#> 641   1.224282526      0
#> 642  -1.365398847      0
#> 643  -0.554276130      0
#> 644  -0.438339628      0
#> 645   1.085859057      0
#> 646   1.316970088      0
#> 647   0.699224321      0
#> 648   0.369221055      0
#> 649  -1.296878062      0
#> 650   0.404045125      0
#> 651  -1.273909017      0
#> 652  -1.030925342      0
#> 653   0.113038288      0
#> 654  -1.776308362      0
#> 655  -1.636060771      0
#> 656   1.078842453      0
#> 657   1.380207911      0
#> 658  -0.040870965      0
#> 659  -0.944308086      0
#> 660   0.605586744      0
#> 661   0.993537609      0
#> 662   0.340471540      0
#> 663  -0.046888344      0
#> 664   0.212118924      0
#> 665   0.133875995      0
#> 666  -0.537486316      0
#> 667  -0.299228328      0
#> 668   0.556925289      0
#> 669   0.825926572      0
#> 670   0.907889687      0
#> 671   0.350122884      0
#> 672  -1.555009179      0
#> 673   0.506893121      0
#> 674  -0.214653201      0
#> 675   1.765475262      0
#> 676   0.623517797      0
#> 677  -0.751027095      0
#> 678  -0.259276673      0
#> 679  -1.169796655      0
#> 680  -1.747189624      0
#> 681   2.973310075      0
#> 682  -0.727192851      0
#> 683  -0.814838590      0
#> 684   0.963178235      0
#> 685   1.460309404      0
#> 686   0.154942235      0
#> 687   2.502701806      0
#> 688   0.070732256      0
#> 689  -0.369952256      0
#> 690  -0.873356383      0
#> 691   0.603692955      0
#> 692  -0.643600984      0
#> 693  -1.458882802      0
#> 694   0.131560807      0
#> 695   0.771674072      0
#> 696  -0.485816534      0
#> 697   2.859981685      0
#> 698   0.545056316      0
#> 699   0.562785533      0
#> 700   1.514899327      0
#> 701   1.037418981      0
#> 702   0.780535824      0
#> 703   1.097146806      0
#> 704   1.730852585      0
#> 705  -0.350536779      0
#> 706  -0.951937072      0
#> 707   0.253981588      0
#> 708   0.049812808      0
#> 709   1.951752409      0
#> 710   0.672563679      0
#> 711   0.167544147      0
#> 712   0.114105768      0
#> 713  -1.143113115      0
#> 714   1.162031592      0
#> 715  -0.371349627      0
#> 716   1.911150487      0
#> 717   0.984115539      0
#> 718  -0.727454939      0
#> 719   1.034778302      0
#> 720   0.811039732      0
#> 721   1.477185599      0
#> 722   0.187466051      0
#> 723   0.391474106      0
#> 724  -0.041735513      0
#> 725   0.177347049      0
#> 726  -2.476331147      0
#> 727  -0.945862964      0
#> 728   0.717156381      0
#> 729   0.609088802      0
#> 730   0.334860948      0
#> 731  -0.253058571      0
#> 732   0.509688925      0
#> 733  -0.984016837      0
#> 734   0.281969880      0
#> 735  -0.981323164      0
#> 736  -1.125165625      0
#> 737   0.264473121      0
#> 738   0.257949382      0
#> 739  -0.336142221      0
#> 740   0.172701352      0
#> 741  -0.778571009      0
#> 742   0.499962590      0
#> 743  -0.352793903      0
#> 744  -1.245140606      0
#> 745  -0.300331679      0
#> 746  -1.764195746      0
#> 747  -1.061246874      0
#> 748   0.318538731      0
#> 749   0.377061862      0
#> 750   0.495169836      0
#> 751   1.161602244      0
#> 752   1.689616580      0
#> 753   0.170739857      0
#> 754  -2.539265539      0
#> 755  -0.307517547      0
#> 756  -0.458822994      0
#> 757   1.281731053      0
#> 758   0.276030980      0
#> 759   0.635810619      0
#> 760  -0.147001073      0
#> 761  -0.624735568      0
#> 762  -0.705367424      0
#> 763  -1.029934613      0
#> 764   0.289533454      0
#> 765   0.837935489      0
#> 766   0.236529068      0
#> 767   0.908446158      0
#> 768   0.848030692      0
#> 769   0.267892892      0
#> 770  -0.409737806      0
#> 771  -0.337242323      0
#> 772  -0.485390735      0
#> 773   0.419881740      0
#> 774   0.505562155      0
#> 775   1.690256133      0
#> 776   0.006610125      0
#> 777   0.342249014      0
#> 778   0.206560804      0
#> 779   0.029672777      0
#> 780   0.118954270      0
#> 781   0.049747975      0
#> 782   0.471011353      0
#> 783  -0.551963187      0
#> 784  -0.266675015      0
#> 785   0.008932740      0
#> 786  -0.461152154      0
#> 787  -0.886423465      0
#> 788  -0.536288030      0
#> 789  -1.176438937      0
#> 790  -0.343965340      0
#> 791  -0.335599597      0
#> 792  -0.940223963      0
#> 793   1.130750769      0
#> 794  -0.172166672      0
#> 795   2.074282093      0
#> 796  -0.105250709      0
#> 797  -1.826782688      0
#> 798  -0.764074278      0
#> 799   0.275566799      0
#> 800  -0.978431932      0
#> 801   0.402585520      0
#> 802  -1.053012622      0
#> 803  -0.099828192      0
#> 804  -2.026669057      0
#> 805  -0.835851339      0
#> 806  -0.490992008      0
#> 807  -1.646345508      0
#> 808  -1.354216677      0
#> 809  -0.396471534      0
#> 810   1.037479112      0
#> 811   0.531213792      0
#> 812   1.250589003      0
#> 813  -1.548544588      0
#> 814  -0.583640679      0
#> 815   1.889678484      0
#> 816  -0.966677937      0
#> 817  -2.076501916      0
#> 818   0.167257939      0
#> 819   0.239781624      0
#> 820  -1.074047788      0
#> 821  -0.588032317      0
#> 822   0.971906277      0
#> 823   1.158540372      0
#> 824   0.498673031      0
#> 825  -1.104162860      0
#> 826   2.016269843      0
#> 827  -0.383567732      0
#> 828   0.498824678      0
#> 829  -1.124877590      0
#> 830  -0.746831998      0
#> 831   0.439127346      0
#> 832   0.305441111      0
#> 833   1.232054581      0
#> 834  -0.247932304      0
#> 835  -0.775878207      0
#> 836  -1.406020972      0
#> 837  -0.778029857      0
#> 838   0.152419429      0
#> 839  -0.772803863      0
#> 840   1.103024163      0
#> 841  -3.510818817      0
#> 842   0.512062752      0
#> 843   0.044422841      0
#> 844  -0.425835312      0
#> 845  -0.238996653      0
#> 846   0.910631950      0
#> 847  -0.309362950      0
#> 848   0.274449442      0
#> 849   0.507233208      0
#> 850   0.986772421      0
#> 851   0.585837490      0
#> 852   0.655195324      0
#> 853  -0.451994679      0
#> 854  -1.006470610      0
#> 855  -0.059482569      0
#> 856  -1.232420791      0
#> 857  -0.537218077      0
#> 858  -0.463170955      0
#> 859  -1.796070873      0
#> 860  -1.390110563      0
#> 861   0.757429565      0
#> 862  -1.432121587      0
#> 863  -0.521964916      0
#> 864  -1.369639689      0
#> 865  -0.297083460      0
#> 866   0.732505748      0
#> 867  -2.099355489      0
#> 868   0.483266328      0
#> 869   0.358334316      0
#> 870   0.533354326      0
#> 871   0.345827276      0
#> 872  -0.518850918      0
#> 873  -0.476824290      0
#> 874  -0.970234989      0
#> 875  -0.508676693      0
#> 876  -0.723554152      0
#> 877  -0.160165978      0
#> 878   1.468417818      0
#> 879   0.422624371      0
#> 880   0.413766232      0
#> 881   1.192723253      0
#> 882  -1.582707591      0
#> 883   1.966753810      0
#> 884  -0.806080468      0
#> 885  -2.692790192      0
#> 886  -0.286416386      0
#> 887  -0.079766988      0
#> 888  -1.258551236      0
#> 889   0.727749649      0
#> 890  -0.952222542      0
#> 891   0.524670873      0
#> 892   0.064577968      0
#> 893   1.292057654      1
#> 894   0.729058953      1
#> 895  -0.867507609      1
#> 896   0.891566632      1
#> 897   2.560961768      1
#> 898   2.006430451      1
#> 899   0.030717467      1
#> 900   0.164905830      1
#> 901   0.066509698      1
#> 902   0.156838917      1
#> 903   0.597338100      1
#> 904   3.279090091      1
#> 905  -0.020500883      1
#> 906   0.527718875      1
#> 907  -0.714017894      1
#> 908   0.679569158      1
#> 909   1.736171909      1
#> 910  -0.167040401      1
#> 911   0.607895423      1
#> 912   0.036360336      1
#> 913   2.172058562      1
#> 914   1.474488894      1
#> 915   0.081329808      1
#> 916  -1.271169384      1
#> 917   0.481298198      1
#> 918   0.237610164      1
#> 919   0.583610948      1
#> 920   0.333680615      1
#> 921   0.055136366      1
#> 922   2.424715695      1
#> 923   0.383515739      1
#> 924   0.550722171      1
#> 925   0.169262843      1
#> 926   1.195297540      1
#> 927   0.105209577      1
#> 928   1.723940337      1
#> 929   0.597410554      1
#> 930   0.965375342      1
#> 931   0.206559021      1
#> 932   0.689555644      1
#> 933   0.337405936      1
#> 934   1.698101690      1
#> 935   2.182882735      1
#> 936   0.515853309      1
#> 937   0.294734802      1
#> 938   1.368489018      1
#> 939   1.746107673      1
#> 940   0.626721383      1
#> 941   1.362510480      1
#> 942   0.100828769      1
#> 943   1.879780531      1
#> 944   1.315202155      1
#> 945   1.532006148      1
#> 946   0.331902899      1
#> 947   0.570801366      1
#> 948  -0.227395872      1
#> 949   1.373331763      1
#> 950  -0.328134802      1
#> 951   1.314545066      1
#> 952   0.999591066      1
#> 953   0.698577169      1
#> 954   0.761293898      1
#> 955  -0.940573459      1
#> 956   0.643248243      1
#> 957   1.615279837      1
#> 958   1.852248708      1
#> 959  -0.575908080      1
#> 960   0.159309676      1
#> 961   1.082114412      1
#> 962   0.635830487      1
#> 963   1.379767569      1
#> 964  -1.307431324      1
#> 965   0.153904879      1
#> 966   1.545288533      1
#> 967   0.774608606      1
#> 968   0.726797009      1
#> 969   0.711076489      1
#> 970  -0.540411544      1
#> 971   0.866579449      1
#> 972  -2.057944264      1
#> 973   0.643219841      1
#> 974   1.253073890      1
#> 975  -1.927808160      1
#> 976   0.864045426      1
#> 977   0.572112750      1
#> 978  -0.116682686      1
#> 979   2.491577210      1
#> 980   0.379584464      1
#> 981   0.272923025      1
#> 982   0.958002498      1
#> 983   0.154864375      1
#> 984   1.205956688      1
#> 985   1.737888922      1
#> 986   0.701369551      1
#> 987   0.502661668      1
#> 988   1.582613187      1
#> 989   0.404058036      1
#> 990   1.718589731      1
#> 991   1.878611426      1
#> 992   1.174701834      1
#> 993   0.592902414      1
#> 994   0.577030885      1
#> 995   0.449236382      1
#> 996   1.212532935      1
#> 997   2.226678240      1
#> 998   2.088800874      1
#> 999   1.727845653      1
#> 1000  3.759128486      1