Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     0.3250004885      0
#> 2     0.3558122039      0
#> 3    -0.0286331299      0
#> 4     0.1694149535      0
#> 5    -0.8275771519      0
#> 6    -0.3291437013      0
#> 7     0.4982748203      0
#> 8    -0.7341735547      0
#> 9    -0.4428272031      0
#> 10    1.4833539631      0
#> 11    0.2762330009      0
#> 12    0.0274149197      0
#> 13   -1.0237514495      0
#> 14   -0.4346354272      0
#> 15   -0.7789366101      0
#> 16   -0.2100877622      0
#> 17   -0.7590062363      0
#> 18   -0.8839707502      0
#> 19    1.3664350663      0
#> 20    0.0667001635      0
#> 21   -0.7425396174      0
#> 22   -0.3199482080      0
#> 23   -0.2549499656      0
#> 24   -1.1159209036      0
#> 25    0.7111074032      0
#> 26   -0.9285006970      0
#> 27   -1.2168973246      0
#> 28    0.6369099775      0
#> 29   -0.0491613689      0
#> 30    0.5776001103      0
#> 31   -0.5611309897      0
#> 32   -2.2588071104      0
#> 33   -0.2379117849      0
#> 34   -0.6479122053      0
#> 35   -0.3450634195      0
#> 36   -1.0034218766      0
#> 37   -1.0051422083      0
#> 38   -2.7088426027      0
#> 39   -0.6128312669      0
#> 40    0.1473004638      0
#> 41   -0.2524435738      0
#> 42   -0.5249222402      0
#> 43    1.3236066293      0
#> 44    1.7659946095      0
#> 45    0.0614805270      0
#> 46    0.5215002421      0
#> 47    0.9496674961      0
#> 48   -1.2236404323      0
#> 49   -0.8767796081      0
#> 50    0.7924440769      0
#> 51   -1.2612153976      0
#> 52    1.4828379844      0
#> 53    0.2827843443      0
#> 54    1.0194278445      0
#> 55   -0.5631130241      0
#> 56   -0.3391225714      0
#> 57    2.1975980683      0
#> 58   -1.3949523117      0
#> 59   -0.2800738620      0
#> 60   -0.0574990956      0
#> 61   -1.1569619104      0
#> 62   -0.8896543430      0
#> 63    1.2277874172      0
#> 64   -0.4084299391      0
#> 65   -0.6562115193      0
#> 66   -1.2452545142      0
#> 67    1.0047139623      0
#> 68    0.0763239615      0
#> 69    1.0562892204      0
#> 70    0.4623393955      0
#> 71    0.9555214272      0
#> 72   -0.4661596534      0
#> 73   -0.0079413447      0
#> 74   -1.3818386578      0
#> 75   -0.4425080415      0
#> 76    1.4431583304      0
#> 77   -0.9821292531      0
#> 78   -0.5732072048      0
#> 79    1.3966291518      0
#> 80   -1.2486018607      0
#> 81    0.7789896609      0
#> 82    0.4771022386      0
#> 83   -0.7246022312      0
#> 84   -0.2748964269      0
#> 85   -0.4861687591      0
#> 86    0.1292733613      0
#> 87   -1.6354942266      0
#> 88    0.5696031094      0
#> 89   -0.2571022307      0
#> 90   -1.1567582125      0
#> 91    1.5949221840      0
#> 92    0.4437880406      0
#> 93    0.5230446928      0
#> 94   -0.4604316347      0
#> 95    0.7931540545      0
#> 96    0.0927341141      0
#> 97    0.2857978645      0
#> 98   -0.2944649072      0
#> 99    0.0552411499      0
#> 100   0.7027981535      0
#> 101  -1.0893818341      0
#> 102  -1.8334185416      0
#> 103  -0.9888685639      0
#> 104  -0.2349233424      0
#> 105  -2.0661162572      0
#> 106  -0.8769944290      0
#> 107   0.6402475652      0
#> 108   0.3646438605      0
#> 109   0.8384158856      0
#> 110   0.8920576547      0
#> 111   1.4029688547      0
#> 112  -1.4395076033      0
#> 113  -0.5341400550      0
#> 114  -1.4758289544      0
#> 115   0.7016484569      0
#> 116  -0.7989741559      0
#> 117   0.6748477422      0
#> 118   0.8538861504      0
#> 119  -0.6689606347      0
#> 120   0.6497485515      0
#> 121  -0.2032985736      0
#> 122   0.2581333827      0
#> 123  -0.5964341564      0
#> 124  -2.0401241152      0
#> 125  -0.1592478081      0
#> 126   1.9566099694      0
#> 127   1.2204727532      0
#> 128  -1.2090115156      0
#> 129  -0.0894224669      0
#> 130   1.0912079101      0
#> 131   0.3172194772      0
#> 132   0.7565262518      0
#> 133  -1.5008228453      0
#> 134   0.2352811211      0
#> 135  -1.8135673528      0
#> 136  -1.3793965717      0
#> 137   1.3503975029      0
#> 138  -2.0093810868      0
#> 139  -1.6177824811      0
#> 140  -0.9893897649      0
#> 141   0.3821799791      0
#> 142  -1.9042280154      0
#> 143   0.3708639536      0
#> 144  -0.7923082343      0
#> 145   0.5954576447      0
#> 146   2.6664801339      0
#> 147  -0.8577164149      0
#> 148   1.1980444596      0
#> 149   0.3895278724      0
#> 150   1.2220073835      0
#> 151  -0.9002050333      0
#> 152  -1.4490230315      0
#> 153  -1.1441736441      0
#> 154  -0.2015581190      0
#> 155  -0.1297879339      0
#> 156  -1.2209535905      0
#> 157  -0.0546692663      0
#> 158  -1.1058218171      0
#> 159  -0.8351605233      0
#> 160   0.8487342853      0
#> 161  -2.2511359865      0
#> 162   0.6332379164      0
#> 163  -0.3946226811      0
#> 164  -0.9980028486      0
#> 165  -1.0949355584      0
#> 166   0.3656167380      0
#> 167  -0.2100556444      0
#> 168   0.8029430623      0
#> 169  -0.8078377424      0
#> 170   1.0443926367      0
#> 171  -0.0745496123      0
#> 172  -1.1995483312      0
#> 173  -0.8547016501      0
#> 174  -1.2476021975      0
#> 175   1.4025380782      0
#> 176   0.6349228558      0
#> 177   0.0314036927      0
#> 178   0.0304673944      0
#> 179  -0.8516684179      0
#> 180   1.3091300840      0
#> 181  -0.4983278919      0
#> 182  -1.2715246766      0
#> 183  -1.2458938783      0
#> 184   0.2670116931      0
#> 185  -1.1623850407      0
#> 186   0.2320473884      0
#> 187  -0.5509282512      0
#> 188  -0.1076288603      0
#> 189  -0.0252298787      0
#> 190  -0.7078509896      0
#> 191  -0.3192865265      0
#> 192   0.3210392922      0
#> 193  -0.0135523062      0
#> 194  -0.2513863136      0
#> 195   1.0553414018      0
#> 196  -0.1121905114      0
#> 197  -0.2934612643      0
#> 198   0.2785270207      0
#> 199   0.7729997025      0
#> 200   1.5629217268      0
#> 201  -0.3524088758      0
#> 202   0.7377979126      0
#> 203   0.7858250939      0
#> 204  -1.1771800624      0
#> 205   0.4181613588      0
#> 206  -0.6729413853      0
#> 207  -0.3630297104      0
#> 208  -0.0385681900      0
#> 209  -1.4938303174      0
#> 210  -0.9798727447      0
#> 211  -2.3415710088      0
#> 212   0.1525694882      0
#> 213   0.1651984582      0
#> 214   0.6996978142      0
#> 215   0.0349898293      0
#> 216  -1.3971310265      0
#> 217  -0.4217193270      0
#> 218  -0.1589907357      0
#> 219   2.1453301217      0
#> 220   2.6095877522      0
#> 221   0.5579345122      0
#> 222  -0.3825678020      0
#> 223  -0.6664047045      0
#> 224  -0.9848086550      0
#> 225   0.4377644554      0
#> 226  -1.0957485185      0
#> 227   0.0975041077      0
#> 228   0.6313484080      0
#> 229  -0.2046444984      0
#> 230  -0.9623755512      0
#> 231  -0.5554260720      0
#> 232  -1.4499614211      0
#> 233   0.8367799635      0
#> 234   0.2339900188      0
#> 235  -0.4662532885      0
#> 236   0.4066721679      0
#> 237   0.1166560604      0
#> 238  -0.2977741279      0
#> 239  -1.8710605689      0
#> 240  -0.1946972032      0
#> 241  -1.6514824522      0
#> 242   1.0930506893      0
#> 243   1.1898197904      0
#> 244   0.0356669102      0
#> 245   1.9709246428      0
#> 246   0.9338527243      0
#> 247   0.9998648274      0
#> 248  -0.1211913699      0
#> 249   0.5871835864      0
#> 250   1.3440918285      0
#> 251  -1.4077421232      0
#> 252   0.5557385814      0
#> 253   1.5872505022      0
#> 254   1.6988173815      0
#> 255  -0.7713884175      0
#> 256  -0.2565737570      0
#> 257  -0.2077133811      0
#> 258   0.7090192090      0
#> 259  -0.4320945413      0
#> 260   0.8555982729      0
#> 261   0.2842318508      0
#> 262   0.4388866616      0
#> 263   0.2197676045      0
#> 264   1.2329002114      0
#> 265   0.1778095709      0
#> 266  -0.5564239459      0
#> 267  -0.3447918121      0
#> 268   0.6878088943      0
#> 269   0.2930960470      0
#> 270   0.8757953435      0
#> 271  -1.2533171078      0
#> 272  -0.7598502812      0
#> 273   0.3433552591      0
#> 274   0.1785589252      0
#> 275   0.8625743037      0
#> 276   0.3863918651      0
#> 277  -0.5425629096      0
#> 278  -0.1002479325      0
#> 279   0.6900433867      0
#> 280  -1.5336734371      0
#> 281   0.2905173276      0
#> 282   0.2324960298      0
#> 283  -1.5424978697      0
#> 284   0.5775456880      0
#> 285  -0.1887061350      0
#> 286   0.7328112536      0
#> 287  -0.1083945796      0
#> 288  -1.1560478261      0
#> 289   1.5661806639      0
#> 290  -1.2763350395      0
#> 291  -1.6832888168      0
#> 292   1.1108009876      0
#> 293  -0.9032694773      0
#> 294  -0.6317547044      0
#> 295   1.3718247556      0
#> 296   0.6739935675      0
#> 297  -0.6971608688      0
#> 298   0.9091844534      0
#> 299  -0.1655725825      0
#> 300   0.5768222761      0
#> 301  -0.8289486468      0
#> 302  -2.6805969406      0
#> 303   0.3613081103      0
#> 304   0.8437536936      0
#> 305  -0.5087632616      0
#> 306   1.3591703265      0
#> 307   0.3983911054      0
#> 308  -0.7553182421      0
#> 309  -0.2743235682      0
#> 310  -0.8162154211      0
#> 311   0.0212704694      0
#> 312  -0.1061096145      0
#> 313  -0.1498227352      0
#> 314  -1.1504357146      0
#> 315  -0.2992407887      0
#> 316   0.6637996309      0
#> 317  -0.7539438462      0
#> 318   0.9989571942      0
#> 319   0.6472584842      0
#> 320   0.1793242980      0
#> 321   1.5232436222      0
#> 322   1.9142056341      0
#> 323   1.4091356356      0
#> 324  -1.2766619864      0
#> 325  -0.5412234451      0
#> 326  -1.2095905630      0
#> 327  -0.9689019117      0
#> 328   2.1421836378      0
#> 329  -1.3269501790      0
#> 330  -1.5789877613      0
#> 331   0.1246052745      0
#> 332  -0.0034015542      0
#> 333  -1.3468899156      0
#> 334   0.3408492894      0
#> 335  -1.1694517800      0
#> 336  -1.2096906567      0
#> 337   0.6975978798      0
#> 338   0.8433774199      0
#> 339   0.2753770279      0
#> 340  -0.3245009628      0
#> 341  -0.5371472440      0
#> 342   1.5449316997      0
#> 343   0.3656387126      0
#> 344   1.0572198133      0
#> 345  -0.2076222707      0
#> 346  -1.5298079680      0
#> 347  -2.4308997089      0
#> 348   1.6676144748      0
#> 349   0.3452747557      0
#> 350  -1.9835015697      0
#> 351   0.7727220305      0
#> 352   0.4562445922      0
#> 353  -0.8109949896      0
#> 354   1.2560649960      0
#> 355  -0.2624306775      0
#> 356  -0.5687356128      0
#> 357   1.8809269697      0
#> 358   0.3064633770      0
#> 359   1.0325584853      0
#> 360   0.1175448762      0
#> 361   1.1413778405      0
#> 362   1.8633957195      0
#> 363  -0.0904200597      0
#> 364   1.8189157302      0
#> 365  -1.2397510627      0
#> 366  -0.5873728745      0
#> 367   0.9793942099      0
#> 368  -0.6619110709      0
#> 369   0.1788267825      0
#> 370   1.4176380154      0
#> 371   0.9665387371      0
#> 372  -0.1773430270      0
#> 373  -0.1914608563      0
#> 374  -0.2219996293      0
#> 375   0.0782781899      0
#> 376   0.4448039134      0
#> 377   0.8187031830      0
#> 378   0.5383693609      0
#> 379  -0.4699234691      0
#> 380  -0.1099084556      0
#> 381  -1.3636765481      0
#> 382  -1.4454512156      0
#> 383   0.3576281997      0
#> 384  -0.2799629635      0
#> 385   0.0696338996      0
#> 386   0.3171069708      0
#> 387  -0.8047148915      0
#> 388   0.0986685839      0
#> 389   0.3098969642      0
#> 390  -0.4850642662      0
#> 391   0.7973273010      0
#> 392  -0.1547165329      0
#> 393   1.8618469682      0
#> 394  -0.1175452661      0
#> 395  -0.9119287081      0
#> 396   0.9672117204      0
#> 397   0.5375846300      0
#> 398  -0.0602783935      0
#> 399   0.9913487551      0
#> 400   2.9247036775      0
#> 401  -0.4776967531      0
#> 402   0.0647411214      0
#> 403  -0.0164180685      0
#> 404  -1.4007812181      0
#> 405  -0.4880672021      0
#> 406  -0.6152750878      0
#> 407  -0.0388402099      0
#> 408   1.3770932263      0
#> 409   0.5047853485      0
#> 410   1.8442428171      0
#> 411   0.2509862172      0
#> 412   0.3690943071      0
#> 413  -1.0917390442      0
#> 414  -0.3699511327      0
#> 415  -2.3232956363      0
#> 416  -0.7143232750      0
#> 417   0.3163433755      0
#> 418   0.3778275876      0
#> 419   1.0416038956      0
#> 420  -1.0751209465      0
#> 421  -0.0252089281      0
#> 422  -0.6466802433      0
#> 423  -0.6752173699      0
#> 424  -0.7441942866      0
#> 425  -0.3147724638      0
#> 426   1.9857491820      0
#> 427  -0.3501370412      0
#> 428   1.2443952081      0
#> 429  -0.1630051084      0
#> 430   0.8249210689      0
#> 431  -0.1750157035      0
#> 432   0.8973700490      0
#> 433  -0.1304198823      0
#> 434  -0.8353872846      0
#> 435  -0.2647423988      0
#> 436  -0.1366719139      0
#> 437   0.6158615225      0
#> 438  -0.8131973556      0
#> 439   0.1674250342      0
#> 440  -0.0640151039      0
#> 441  -0.1603625568      0
#> 442   0.5433732049      0
#> 443  -0.3179077896      0
#> 444   0.2788673343      0
#> 445  -1.2163139090      0
#> 446   2.0527062109      0
#> 447  -0.7264372591      0
#> 448  -0.3521409206      0
#> 449   3.0217327638      0
#> 450   0.7645626825      0
#> 451  -0.5235801314      0
#> 452  -0.5636077643      0
#> 453  -1.1190023987      0
#> 454  -0.7946472877      0
#> 455   1.4019816724      0
#> 456  -0.7074876698      0
#> 457   2.3875857603      0
#> 458   0.9171734035      0
#> 459  -1.0464788219      0
#> 460   0.2324014559      0
#> 461  -0.7436685472      0
#> 462   0.5214409741      0
#> 463   1.3762073140      0
#> 464   0.0168403223      0
#> 465   0.9526181786      0
#> 466  -1.1763345061      0
#> 467   0.2287386641      0
#> 468   0.1325417230      0
#> 469  -0.9714052529      0
#> 470  -2.8733084185      0
#> 471  -0.4897260211      0
#> 472   1.2029154472      0
#> 473  -0.4597941561      0
#> 474   1.6400703593      0
#> 475  -1.5614546961      0
#> 476  -0.6285165542      0
#> 477  -1.1937643743      0
#> 478   0.0942013000      0
#> 479   0.1632926550      0
#> 480  -0.5266253463      0
#> 481   1.5518606258      0
#> 482  -0.8221142292      0
#> 483  -0.0286111220      0
#> 484   1.2446053673      0
#> 485  -0.2575663417      0
#> 486  -0.1843641831      0
#> 487   0.3601245865      0
#> 488   1.6456247648      0
#> 489   0.5258505117      0
#> 490   1.6739104184      0
#> 491   1.7097578462      0
#> 492   0.2821652379      0
#> 493   2.1997348526      0
#> 494   1.1037323495      0
#> 495   1.2659669201      0
#> 496  -0.0885322342      0
#> 497  -0.5987597280      0
#> 498   1.9124917661      0
#> 499   0.8231825693      0
#> 500   0.1640204667      0
#> 501  -0.1255526020      0
#> 502   0.6805785802      0
#> 503   1.3327574293      0
#> 504   1.1287004321      0
#> 505   0.1292773515      0
#> 506  -0.2695484185      0
#> 507  -0.2393414725      0
#> 508  -1.8072907101      0
#> 509  -1.2619326525      0
#> 510  -0.7262288717      0
#> 511  -0.3782123134      0
#> 512  -0.3909111251      0
#> 513  -0.2785502151      0
#> 514   1.2269532368      0
#> 515   0.3188749540      0
#> 516  -1.1388098115      0
#> 517  -1.2325375906      0
#> 518  -0.7888677330      0
#> 519   2.2328633506      0
#> 520   0.9606532881      0
#> 521   0.0752511483      0
#> 522   0.5807169032      0
#> 523  -0.2198253998      0
#> 524   2.2205980940      0
#> 525  -1.8563334004      0
#> 526   1.6474485065      0
#> 527  -0.8365891131      0
#> 528  -0.6262897509      0
#> 529   1.1755463034      0
#> 530  -0.8032500989      0
#> 531   2.7345758664      0
#> 532   1.3296012508      0
#> 533   0.2576746534      0
#> 534  -0.7242470111      0
#> 535   1.0547418723      0
#> 536   0.7206040844      0
#> 537  -1.1293008238      0
#> 538  -0.7261079198      0
#> 539  -0.0320684077      0
#> 540   1.3319289819      0
#> 541   1.2578857978      0
#> 542   0.1477145369      0
#> 543  -2.3385595156      0
#> 544  -0.3469955509      0
#> 545  -1.0357658849      0
#> 546   0.1301642392      0
#> 547   0.2586893365      0
#> 548   0.4414547672      0
#> 549  -1.4377012550      0
#> 550   0.5475264611      0
#> 551  -1.1008619369      0
#> 552   1.1261260427      0
#> 553  -1.2195898609      0
#> 554  -0.4969377157      0
#> 555  -0.6887453773      0
#> 556   0.3746433545      0
#> 557   0.8537043722      0
#> 558   1.2741018228      0
#> 559   0.2077830900      0
#> 560   1.4293609877      0
#> 561  -1.5338596246      0
#> 562  -0.1622291000      0
#> 563   0.2032684917      0
#> 564   1.2167276956      0
#> 565  -0.6522886872      0
#> 566   0.2871527169      0
#> 567  -0.0672969444      0
#> 568   0.5941128051      0
#> 569  -0.5623394467      0
#> 570   0.8245970722      0
#> 571  -0.3521311973      0
#> 572  -0.1114581099      0
#> 573  -0.7657796553      0
#> 574  -0.0118046480      0
#> 575  -0.4007205695      0
#> 576   0.7530942297      0
#> 577  -0.6926187069      0
#> 578   0.2819546691      0
#> 579   0.7024631334      0
#> 580  -0.6386431622      0
#> 581   0.1821644876      0
#> 582  -0.2978461751      0
#> 583   0.3338282425      0
#> 584  -1.2584707939      0
#> 585   0.8429075536      0
#> 586  -0.3073017459      0
#> 587  -0.2388258909      0
#> 588   1.5762213321      0
#> 589  -1.0356094226      0
#> 590   0.7369691855      0
#> 591  -0.1962369359      0
#> 592  -0.8375350898      0
#> 593   0.5436796858      0
#> 594  -0.8085770328      0
#> 595   0.2818928656      0
#> 596  -1.3813763009      0
#> 597  -0.2750542419      0
#> 598   0.4288841330      0
#> 599  -0.4603478163      0
#> 600  -0.7017359754      0
#> 601   0.3494697321      0
#> 602   0.1886853955      0
#> 603  -0.7525112768      0
#> 604   2.0674650467      0
#> 605  -1.4625430348      0
#> 606   0.3411302515      0
#> 607   0.4910789258      0
#> 608  -1.6859675537      0
#> 609  -0.0719626469      0
#> 610   0.3332530051      0
#> 611   0.5223597123      0
#> 612  -0.3337880997      0
#> 613  -0.4927656006      0
#> 614   1.3081462340      0
#> 615  -0.6501610210      0
#> 616  -0.6321425802      0
#> 617   0.8722044611      0
#> 618  -0.0770655371      0
#> 619  -0.5401742367      0
#> 620  -0.2494189133      0
#> 621  -0.3977405083      0
#> 622   0.0222774375      0
#> 623  -2.1108405846      0
#> 624  -1.7307002694      0
#> 625  -0.0643823289      0
#> 626   0.5596598257      0
#> 627   1.8544647009      0
#> 628  -0.0119980055      0
#> 629   0.2278193955      0
#> 630  -0.1691266181      0
#> 631   1.5656338305      0
#> 632  -0.9928245672      0
#> 633   0.4373663127      0
#> 634  -0.4588699123      0
#> 635   0.0367073328      0
#> 636   0.1175737980      0
#> 637   0.2954011299      0
#> 638   0.8083498580      0
#> 639   0.1962242327      0
#> 640   1.0904748259      0
#> 641  -0.3776713609      0
#> 642   0.6437430165      0
#> 643  -0.6548712521      0
#> 644   1.6773071424      0
#> 645   1.1572014660      0
#> 646  -0.0493924774      0
#> 647  -0.6208527420      0
#> 648   2.2679890532      0
#> 649  -0.6583487086      0
#> 650  -0.6215666740      0
#> 651  -0.5654921586      0
#> 652  -1.1707762933      0
#> 653   1.2938914837      0
#> 654  -1.0215699638      0
#> 655   0.6381146159      0
#> 656   1.0941618699      0
#> 657   0.2197981826      0
#> 658   2.1141791500      0
#> 659  -0.4449923620      0
#> 660   0.4477036050      0
#> 661   0.3601528561      0
#> 662  -0.8968352589      0
#> 663   1.2753648298      0
#> 664  -0.1249188764      0
#> 665  -1.0818452351      0
#> 666  -0.7005023016      0
#> 667  -0.6460887876      0
#> 668  -0.9843415013      0
#> 669  -0.5531617959      0
#> 670  -0.1106871103      0
#> 671  -0.4749613869      0
#> 672  -0.8149405563      0
#> 673   0.0222755678      0
#> 674   0.4612459310      0
#> 675  -0.7219015479      0
#> 676  -0.4119286449      0
#> 677  -0.3752410969      0
#> 678  -0.1752564832      0
#> 679   1.6538808123      0
#> 680  -0.7382917193      0
#> 681  -0.0992485621      0
#> 682   0.0489822219      0
#> 683  -2.7146513314      0
#> 684   0.5933746029      0
#> 685   1.3995093029      0
#> 686   0.0487159569      0
#> 687   0.3407053933      0
#> 688  -0.5037843871      0
#> 689   1.1400735417      0
#> 690   1.0145920220      0
#> 691   1.1565374728      0
#> 692   1.2403643368      0
#> 693   0.6556775944      0
#> 694   0.2038405564      0
#> 695   0.0918603218      0
#> 696   1.5441139910      0
#> 697  -0.9649167340      0
#> 698   1.8141959821      0
#> 699  -1.3303277264      0
#> 700  -1.3150871656      0
#> 701  -0.3335209568      0
#> 702  -1.2563984266      0
#> 703   1.5603348802      0
#> 704  -0.6715879850      0
#> 705   0.6089641919      0
#> 706  -0.5016272252      0
#> 707  -0.9251092656      0
#> 708   0.0890303871      0
#> 709  -0.4796790215      0
#> 710  -0.1258356577      0
#> 711   0.1690506881      0
#> 712   2.5078826091      0
#> 713  -0.2067595597      0
#> 714  -0.1545097301      0
#> 715  -0.5012446881      0
#> 716   1.1448574802      0
#> 717   0.6895596570      0
#> 718   0.0449083905      0
#> 719   0.1509293450      0
#> 720   0.2446053220      0
#> 721  -0.4862673075      0
#> 722  -1.8996153751      0
#> 723   0.5644843029      0
#> 724  -1.7634781402      0
#> 725  -0.5183555693      0
#> 726  -0.0809337940      0
#> 727  -0.5121479397      0
#> 728  -3.1668666636      0
#> 729   0.5804732707      0
#> 730  -1.6987330861      0
#> 731  -0.5683631026      0
#> 732  -0.7223510922      0
#> 733  -1.3209240153      0
#> 734  -0.1580048332      0
#> 735  -2.0479189613      0
#> 736   0.1701403948      0
#> 737   0.0577476169      0
#> 738   0.0520794987      0
#> 739  -0.7938774672      0
#> 740   1.6340014535      0
#> 741   0.4993511659      0
#> 742   0.1310508356      0
#> 743   1.7390737127      0
#> 744  -0.4359950745      0
#> 745   1.3728357906      0
#> 746  -0.0001926614      0
#> 747   1.2541804650      0
#> 748  -0.4964490209      0
#> 749   0.3997688486      0
#> 750   0.1253087582      0
#> 751  -1.0422281492      0
#> 752   0.0835779121      0
#> 753   0.0950396203      0
#> 754  -1.6880125130      0
#> 755  -0.8255598685      0
#> 756   1.0765439402      0
#> 757  -0.1717607293      0
#> 758  -1.1649350301      0
#> 759   0.2647161774      0
#> 760  -0.7220125174      0
#> 761  -0.2563790339      0
#> 762   1.1117421898      0
#> 763   0.7445486579      0
#> 764  -0.3343433995      0
#> 765   0.6095698408      0
#> 766   0.8084776752      0
#> 767   0.4373990183      0
#> 768   0.0926257857      0
#> 769  -1.8318490973      0
#> 770  -0.0244850764      0
#> 771  -0.6740740033      0
#> 772  -1.1800918069      0
#> 773   1.1817780790      0
#> 774   0.9648998759      0
#> 775  -0.5894668306      0
#> 776  -0.2356763877      0
#> 777   0.6273926416      0
#> 778   1.3244676542      0
#> 779   0.1361634971      0
#> 780  -0.0826696864      0
#> 781  -0.7540029212      0
#> 782  -0.6176935367      0
#> 783   0.9839782880      0
#> 784   0.0972615572      0
#> 785   0.0468447658      0
#> 786   0.0091886706      0
#> 787  -1.1237934980      0
#> 788   0.0958920837      0
#> 789  -0.2965813817      0
#> 790   0.6916167786      0
#> 791   0.5412941815      0
#> 792   0.6484558103      0
#> 793  -1.3274661903      0
#> 794   0.0878864426      0
#> 795  -0.7614182647      0
#> 796   2.5050534919      0
#> 797   0.5156717815      0
#> 798  -1.0214709547      0
#> 799  -1.5419889430      0
#> 800  -0.1606582232      0
#> 801  -1.2367509853      0
#> 802  -0.3695800803      0
#> 803  -1.0062348490      0
#> 804  -1.1956729884      0
#> 805  -0.9105095438      0
#> 806   0.4477940564      0
#> 807  -1.2163571571      0
#> 808  -0.4474561215      0
#> 809  -0.8932221930      0
#> 810   1.5523965535      0
#> 811   1.1008412180      0
#> 812  -0.7751513701      0
#> 813   0.4341423089      0
#> 814   1.5823070551      0
#> 815   0.4682852948      0
#> 816   1.6730885462      0
#> 817   0.3371455286      0
#> 818  -0.2391918631      0
#> 819  -0.8545849093      0
#> 820   0.5051717922      0
#> 821  -0.4532942710      0
#> 822  -0.2053355223      0
#> 823   0.0255902005      0
#> 824   0.1300141720      0
#> 825   0.4180011705      0
#> 826   1.1023274532      0
#> 827  -0.4197138500      0
#> 828  -1.0111391999      0
#> 829  -0.4124546011      0
#> 830  -0.4985706103      0
#> 831  -0.2980665422      0
#> 832  -0.0446688153      0
#> 833   1.3925045121      0
#> 834   0.0083075990      0
#> 835  -0.6607685838      0
#> 836   1.0096821310      0
#> 837   0.2898184554      0
#> 838  -2.1616407386      0
#> 839  -0.7848838721      0
#> 840  -0.0401193965      0
#> 841   0.2417934586      0
#> 842   1.3685269511      0
#> 843  -0.1636938738      0
#> 844   0.5373126219      0
#> 845  -0.9750571522      0
#> 846  -1.0272131165      0
#> 847   0.1778244416      0
#> 848   0.2749245751      0
#> 849   0.0910054791      0
#> 850   1.3021937218      0
#> 851   1.6222122086      0
#> 852  -0.7076590558      0
#> 853  -1.3996219000      0
#> 854  -0.0044159477      0
#> 855  -0.0007427375      0
#> 856  -0.3274156892      0
#> 857  -1.6222459559      0
#> 858  -0.2702229768      0
#> 859   0.8267776513      0
#> 860   0.7761488957      0
#> 861   0.0102968594      0
#> 862   0.0426069663      0
#> 863  -0.6910333334      0
#> 864   0.5421270252      0
#> 865  -0.6411208147      0
#> 866  -0.6487614710      0
#> 867  -1.5333774765      0
#> 868  -2.3218350734      0
#> 869   0.6390496997      0
#> 870   1.0106689512      0
#> 871   0.1393227853      0
#> 872  -0.2708291070      0
#> 873   1.7848444818      0
#> 874  -1.1165552918      0
#> 875   1.3443399460      0
#> 876  -1.1351468492      0
#> 877  -1.0399169000      0
#> 878   1.4207426682      0
#> 879   1.1801711258      0
#> 880   1.0068189807      0
#> 881   0.2396026154      0
#> 882   0.3769609020      0
#> 883   0.5054350411      0
#> 884  -0.2151261891      0
#> 885  -0.6720281040      0
#> 886   0.8702354399      0
#> 887  -2.3191354081      0
#> 888  -0.0409977874      0
#> 889   1.3594395409      0
#> 890   0.2741286254      0
#> 891  -0.6316484115      0
#> 892  -0.8595649817      0
#> 893  -0.8884293086      0
#> 894   1.0474281848      0
#> 895  -1.1546194713      0
#> 896  -0.0605240470      0
#> 897  -0.1580248601      0
#> 898  -1.7935168044      0
#> 899   0.6631774441      1
#> 900   0.4930422012      1
#> 901   1.2204298030      1
#> 902  -0.3604536966      1
#> 903   0.9117712876      1
#> 904   1.1258188659      1
#> 905   0.4446967034      1
#> 906   0.1830224389      1
#> 907   0.5489245936      1
#> 908   1.1255575664      1
#> 909  -0.1534876566      1
#> 910   0.0336567466      1
#> 911  -0.9992905523      1
#> 912  -0.4508562531      1
#> 913   1.5212259895      1
#> 914   0.3108844553      1
#> 915   0.5307813660      1
#> 916   1.4009222627      1
#> 917   1.4017808168      1
#> 918   1.6731419471      1
#> 919   1.8210052809      1
#> 920  -1.2929800278      1
#> 921   1.5918442556      1
#> 922  -1.7524018316      1
#> 923   1.5705732869      1
#> 924   0.3066300143      1
#> 925  -0.1432804941      1
#> 926   0.2853082396      1
#> 927   0.1668265747      1
#> 928   0.8295073813      1
#> 929   2.0493822624      1
#> 930  -1.3451895739      1
#> 931   2.8562946900      1
#> 932   0.1258960191      1
#> 933   2.3284774761      1
#> 934   2.0350382480      1
#> 935  -0.5859831035      1
#> 936   0.2675053350      1
#> 937   0.0682287725      1
#> 938  -0.6609247815      1
#> 939   0.0450839044      1
#> 940   0.8109195738      1
#> 941   1.2211798167      1
#> 942  -1.3922961057      1
#> 943   0.6733934493      1
#> 944  -0.2020010287      1
#> 945  -0.1327103784      1
#> 946   0.9935533951      1
#> 947  -0.6691377319      1
#> 948   1.1083845241      1
#> 949   1.5443431095      1
#> 950   1.7905081846      1
#> 951   1.1429062850      1
#> 952   1.5082000877      1
#> 953   0.5815352925      1
#> 954  -0.4851218169      1
#> 955  -0.5192868931      1
#> 956   1.1092445354      1
#> 957   1.7592535629      1
#> 958   2.8721930157      1
#> 959   0.6449882503      1
#> 960   2.3306460542      1
#> 961   0.1979723930      1
#> 962   1.5617697732      1
#> 963   0.8768083875      1
#> 964   0.4280761851      1
#> 965   1.0729441364      1
#> 966   0.2255036330      1
#> 967   0.7508481412      1
#> 968   0.9033988730      1
#> 969   1.1121824831      1
#> 970  -0.3379343965      1
#> 971   0.4713063952      1
#> 972  -0.0599740101      1
#> 973   0.9597342954      1
#> 974   1.5300702199      1
#> 975  -1.0455596966      1
#> 976  -1.5791897641      1
#> 977  -0.4333708041      1
#> 978   2.0137307929      1
#> 979   0.4122328226      1
#> 980   2.3399901483      1
#> 981   0.6723133986      1
#> 982   1.1400245930      1
#> 983  -1.0049217717      1
#> 984  -0.4257589740      1
#> 985   0.9177908025      1
#> 986   1.3067323943      1
#> 987  -0.0537518020      1
#> 988   1.1715992163      1
#> 989   0.1863567443      1
#> 990   1.5335177665      1
#> 991  -0.7142436917      1
#> 992   1.4295283722      1
#> 993  -0.3490910123      1
#> 994   0.0406013989      1
#> 995  -2.0264179716      1
#> 996   0.6778928976      1
#> 997   2.0829523055      1
#> 998   2.2141909412      1
#> 999   1.3393808257      1
#> 1000  1.0831680447      1