Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1     0.380074861      0
#> 2    -0.184554758      0
#> 3    -2.092036419      0
#> 4     0.926438515      0
#> 5    -0.014154031      0
#> 6    -1.724727292      0
#> 7     0.407647088      0
#> 8    -0.722601340      0
#> 9     0.155025547      0
#> 10    1.899247570      0
#> 11   -0.739187581      0
#> 12   -1.269241597      0
#> 13    0.186615771      0
#> 14   -0.350969758      0
#> 15    0.348788286      0
#> 16   -0.417658506      0
#> 17   -0.494522527      0
#> 18    1.044445110      0
#> 19    0.041294629      0
#> 20   -1.079834768      0
#> 21    0.461503741      0
#> 22    1.723443772      0
#> 23    1.559123582      0
#> 24    0.701421671      0
#> 25   -0.081477938      0
#> 26   -0.422542692      0
#> 27    0.094683619      0
#> 28    0.106112200      0
#> 29   -1.597621550      0
#> 30   -0.784071451      0
#> 31   -0.449019892      0
#> 32    0.890154656      0
#> 33    1.538004870      0
#> 34    0.350262077      0
#> 35    0.322697227      0
#> 36   -0.038536941      0
#> 37    1.776736221      0
#> 38    1.208263350      0
#> 39   -0.210254214      0
#> 40   -0.256754992      0
#> 41   -0.218830096      0
#> 42   -0.171234160      0
#> 43    0.597309536      0
#> 44    0.838667211      0
#> 45   -1.804802949      0
#> 46    0.558005490      0
#> 47    1.391947619      0
#> 48   -1.639166092      0
#> 49   -1.066690538      0
#> 50   -0.639246412      0
#> 51   -0.044710718      0
#> 52    0.949778008      0
#> 53    0.227117609      0
#> 54    1.391714780      0
#> 55    0.523814338      0
#> 56   -0.697594905      0
#> 57   -0.378452877      0
#> 58   -0.811805052      0
#> 59    1.309203524      0
#> 60   -0.285345046      0
#> 61   -1.034705125      0
#> 62    0.001252039      0
#> 63   -0.728079289      0
#> 64    0.253156159      0
#> 65   -1.132304624      0
#> 66   -1.028515807      0
#> 67    0.088162939      0
#> 68    0.315959673      0
#> 69   -0.459366828      0
#> 70   -1.442015914      0
#> 71   -3.228296010      0
#> 72   -0.589870001      0
#> 73   -1.670937230      0
#> 74   -2.722034550      0
#> 75    0.979462898      0
#> 76    1.983878462      0
#> 77   -0.445535683      0
#> 78   -1.353675067      0
#> 79   -0.230159778      0
#> 80   -1.413189281      0
#> 81   -0.465937197      0
#> 82   -0.386500565      0
#> 83    0.553217383      0
#> 84   -1.345382445      0
#> 85   -0.747363572      0
#> 86    1.444379840      0
#> 87    0.029570651      0
#> 88   -1.527643503      0
#> 89   -0.561608198      0
#> 90   -1.108986379      0
#> 91   -2.488564212      0
#> 92    1.227794834      0
#> 93   -0.845421367      0
#> 94    0.735736048      0
#> 95   -2.086032143      0
#> 96    0.059195745      0
#> 97    0.148990489      0
#> 98    0.632816215      0
#> 99    0.111183222      0
#> 100   0.769137104      0
#> 101   2.906343754      0
#> 102   1.033756000      0
#> 103  -1.201746931      0
#> 104   0.016551573      0
#> 105  -0.646389163      0
#> 106  -0.518755336      0
#> 107   0.178740632      0
#> 108  -0.166173098      0
#> 109   2.406510494      0
#> 110   0.260358136      0
#> 111   1.916543286      0
#> 112   0.960545114      0
#> 113   2.098981741      0
#> 114   0.208879297      0
#> 115  -0.032441134      0
#> 116  -0.094619883      0
#> 117  -0.547386738      0
#> 118  -1.145558576      0
#> 119   0.726206541      0
#> 120   0.908624498      0
#> 121   1.724153020      0
#> 122  -0.285793197      0
#> 123   0.498919784      0
#> 124  -1.709590147      0
#> 125  -0.886948883      0
#> 126  -1.354427861      0
#> 127  -2.035781281      0
#> 128  -0.697226886      0
#> 129   0.872034818      0
#> 130   0.014545150      0
#> 131   0.619209625      0
#> 132   0.651432062      0
#> 133  -2.265635921      0
#> 134  -0.157297289      0
#> 135   0.516884944      0
#> 136   0.338889214      0
#> 137   0.099143614      0
#> 138  -0.200327971      0
#> 139   1.030196324      0
#> 140   1.231685754      0
#> 141  -0.579864791      0
#> 142  -1.260720156      0
#> 143  -0.674871317      0
#> 144   1.119427522      0
#> 145  -0.598858199      0
#> 146  -0.935796211      0
#> 147   0.926307710      0
#> 148  -1.503306751      0
#> 149   0.502757502      0
#> 150   0.520238688      0
#> 151  -0.415400596      0
#> 152   0.757573242      0
#> 153   1.142697547      0
#> 154   1.022227726      0
#> 155   0.475893023      0
#> 156   0.464291412      0
#> 157  -0.480212857      0
#> 158  -0.313741471      0
#> 159   0.419738379      0
#> 160   0.491231924      0
#> 161   0.781352490      0
#> 162   1.576485195      0
#> 163  -0.070513232      0
#> 164  -0.759844994      0
#> 165  -0.781117484      0
#> 166  -0.281124285      0
#> 167   0.232427754      0
#> 168  -0.229350749      0
#> 169   1.441010471      0
#> 170  -0.518705726      0
#> 171  -0.611726766      0
#> 172  -0.750815491      0
#> 173  -2.135200547      0
#> 174  -0.918814695      0
#> 175   0.816329629      0
#> 176  -1.137168250      0
#> 177   0.353426947      0
#> 178   0.606129542      0
#> 179   0.076493663      0
#> 180  -0.094094805      0
#> 181  -1.344837173      0
#> 182  -0.934398228      0
#> 183  -1.907313946      0
#> 184   1.432848986      0
#> 185  -1.054001272      0
#> 186  -0.835010540      0
#> 187   1.245136727      0
#> 188  -0.790200136      0
#> 189   0.160942892      0
#> 190  -1.909994393      0
#> 191   2.180495772      0
#> 192   0.701318064      0
#> 193  -0.466527404      0
#> 194  -1.456041948      0
#> 195   1.286931629      0
#> 196   1.260532415      0
#> 197   0.040505732      0
#> 198   0.268281515      0
#> 199   1.208506342      0
#> 200  -0.345508340      0
#> 201  -1.148673269      0
#> 202   0.938291624      0
#> 203   0.240193652      0
#> 204  -0.531142155      0
#> 205  -0.652303165      0
#> 206  -0.550609476      0
#> 207   0.414120485      0
#> 208  -0.962178301      0
#> 209   0.088611776      0
#> 210   1.513055231      0
#> 211   1.788872830      0
#> 212   1.095328317      0
#> 213   1.013770432      0
#> 214   0.516332752      0
#> 215   1.405251943      0
#> 216   2.063033551      0
#> 217   1.109835606      0
#> 218  -0.272896470      0
#> 219   0.655730064      0
#> 220  -0.256037362      0
#> 221   0.093603200      0
#> 222  -0.858195597      0
#> 223  -1.108765087      0
#> 224   0.347606642      0
#> 225   0.070772900      0
#> 226   1.578717190      0
#> 227  -1.545789660      0
#> 228   0.001086755      0
#> 229   0.660322522      0
#> 230  -0.471854529      0
#> 231  -0.045255147      0
#> 232  -0.227211739      0
#> 233   1.090349255      0
#> 234   0.777063456      0
#> 235  -0.823931209      0
#> 236  -1.284090667      0
#> 237  -2.013238577      0
#> 238   0.150367636      0
#> 239   0.454452259      0
#> 240   0.275485393      0
#> 241  -0.252595248      0
#> 242   1.636609200      0
#> 243   0.109494998      0
#> 244  -0.925751542      0
#> 245   0.314602401      0
#> 246   0.168456818      0
#> 247   0.013171083      0
#> 248  -0.424181349      0
#> 249   0.042690697      0
#> 250   0.863996754      0
#> 251   1.031849240      0
#> 252  -0.828653416      0
#> 253   0.401550649      0
#> 254  -0.045297591      0
#> 255   1.450090130      0
#> 256  -0.398635290      0
#> 257  -0.584489540      0
#> 258   0.290339630      0
#> 259  -0.524020234      0
#> 260  -0.590597294      0
#> 261   1.034410558      0
#> 262   2.032764575      0
#> 263   0.177703151      0
#> 264   1.229072874      0
#> 265  -0.262849209      0
#> 266  -0.374712434      0
#> 267   0.890000027      0
#> 268   1.598183731      0
#> 269  -0.708635642      0
#> 270   0.136009810      0
#> 271  -0.315132265      0
#> 272   0.109474528      0
#> 273   0.430044069      0
#> 274  -0.520522751      0
#> 275  -0.394526378      0
#> 276  -1.667818599      0
#> 277   1.656324481      0
#> 278  -0.908888204      0
#> 279  -0.241027245      0
#> 280  -0.467223362      0
#> 281  -1.726522344      0
#> 282  -0.950324989      0
#> 283   0.119392874      0
#> 284   1.140192183      0
#> 285   0.762115859      0
#> 286  -1.259514833      0
#> 287  -0.622653781      0
#> 288  -1.445713686      0
#> 289  -2.404722022      0
#> 290   0.245057060      0
#> 291  -0.738663282      0
#> 292   0.132387438      0
#> 293   1.618666979      0
#> 294  -0.058936473      0
#> 295   0.877212791      0
#> 296  -0.029281146      0
#> 297   0.746205352      0
#> 298   2.168162288      0
#> 299  -0.967942017      0
#> 300   0.161950914      0
#> 301   0.612547810      0
#> 302   0.646624014      0
#> 303  -0.293964876      0
#> 304  -0.127244323      0
#> 305   1.227984766      0
#> 306   0.673317470      0
#> 307  -0.029901348      0
#> 308   0.187846644      0
#> 309   1.665083056      0
#> 310   2.017793041      0
#> 311  -0.601778118      0
#> 312  -0.045761173      0
#> 313   0.319111308      0
#> 314   1.049980178      0
#> 315  -1.047146042      0
#> 316  -1.423197392      0
#> 317   0.956786877      0
#> 318   1.042407846      0
#> 319  -0.058881688      0
#> 320  -0.755239948      0
#> 321  -0.609807823      0
#> 322  -1.428536571      0
#> 323  -0.952898054      0
#> 324  -1.296810573      0
#> 325  -1.391863826      0
#> 326   1.185507569      0
#> 327   0.210736627      0
#> 328   0.852753375      0
#> 329   0.632113981      0
#> 330  -0.452754375      0
#> 331  -1.729497153      0
#> 332  -1.859642107      0
#> 333   0.441587278      0
#> 334  -0.111152295      0
#> 335  -1.781951593      0
#> 336   0.092340756      0
#> 337  -1.085978690      0
#> 338   0.751824116      0
#> 339  -0.603880461      0
#> 340  -0.339007339      0
#> 341  -1.010279410      0
#> 342   0.168071989      0
#> 343   2.502785669      0
#> 344  -0.604750340      0
#> 345   0.566934139      0
#> 346  -2.284576744      0
#> 347   0.161950671      0
#> 348  -0.055021204      0
#> 349   1.367574101      0
#> 350   0.049306179      0
#> 351   0.393167339      0
#> 352  -1.709606945      0
#> 353  -0.450289753      0
#> 354   1.827090918      0
#> 355  -0.855832867      0
#> 356  -0.988683159      0
#> 357   1.037710576      0
#> 358   0.893863911      0
#> 359  -1.667552869      0
#> 360  -0.783920230      0
#> 361   0.566496874      0
#> 362   0.069001195      0
#> 363  -0.282302015      0
#> 364  -2.159894235      0
#> 365   0.652664655      0
#> 366  -0.675848753      0
#> 367   1.134421555      0
#> 368   1.360050992      0
#> 369  -0.154460432      0
#> 370  -0.712409677      0
#> 371   1.443345804      0
#> 372  -0.919022968      0
#> 373   0.390360332      0
#> 374  -0.009050263      0
#> 375   0.763613913      0
#> 376   0.115084338      0
#> 377  -0.669336146      0
#> 378  -0.855296050      0
#> 379   0.028540148      0
#> 380   1.635974099      0
#> 381  -1.642303910      0
#> 382   1.045056192      0
#> 383  -0.694258830      0
#> 384   0.922627850      0
#> 385   0.277294618      0
#> 386   0.574160698      0
#> 387   0.302731562      0
#> 388   1.656949271      0
#> 389   0.811746029      0
#> 390   0.690753485      0
#> 391   1.104118317      0
#> 392   1.614153420      0
#> 393   0.326997551      0
#> 394   0.968901703      0
#> 395   0.566113580      0
#> 396   0.791587877      0
#> 397  -0.132736887      0
#> 398   0.600286119      0
#> 399  -1.311192994      0
#> 400   0.187344498      0
#> 401  -0.837067123      0
#> 402   0.986023354      0
#> 403  -0.154947498      0
#> 404   0.759707594      0
#> 405   0.167849726      0
#> 406   1.275923746      0
#> 407  -0.082098595      0
#> 408  -2.073384416      0
#> 409   1.151155611      0
#> 410   0.215207952      0
#> 411  -0.158398857      0
#> 412   0.289589258      0
#> 413   0.586296110      0
#> 414  -1.991218476      0
#> 415   2.335148879      0
#> 416   0.031913404      0
#> 417   0.114022915      0
#> 418   1.244081577      0
#> 419  -1.505186498      0
#> 420   0.655369071      0
#> 421   0.581413348      0
#> 422   1.153921792      0
#> 423  -0.510716865      0
#> 424   3.260120440      0
#> 425  -2.017373478      0
#> 426  -0.244983555      0
#> 427   0.037999508      0
#> 428   0.076758089      0
#> 429   0.358322934      0
#> 430   0.668818704      0
#> 431   1.426939116      0
#> 432   1.353608276      0
#> 433   0.120791348      0
#> 434  -0.348044224      0
#> 435  -0.526641043      0
#> 436   0.402857890      0
#> 437  -1.744116679      0
#> 438  -0.465503670      0
#> 439   0.226977326      0
#> 440  -1.119357461      0
#> 441   1.740296866      0
#> 442  -0.413408582      0
#> 443   0.320710641      0
#> 444  -1.842702308      0
#> 445  -1.887707497      0
#> 446   0.739478571      0
#> 447  -1.307993093      0
#> 448   0.850208354      0
#> 449  -0.212240514      0
#> 450   0.187264503      0
#> 451   0.804042182      0
#> 452  -0.218549144      0
#> 453   1.359597468      0
#> 454   0.012132525      0
#> 455  -0.711260596      0
#> 456   0.242954788      0
#> 457   2.424172816      0
#> 458   0.700404876      0
#> 459   1.613104143      0
#> 460   1.643490046      0
#> 461  -0.622266968      0
#> 462   0.742450714      0
#> 463   0.882464175      0
#> 464   2.182116588      0
#> 465  -0.325520880      0
#> 466   0.661939367      0
#> 467  -0.463831147      0
#> 468   0.506285709      0
#> 469  -0.966257603      0
#> 470  -1.145638251      0
#> 471   0.195503644      0
#> 472   0.217166962      0
#> 473  -0.136754474      0
#> 474  -1.102740527      0
#> 475  -1.771814801      0
#> 476   0.987762127      0
#> 477   1.326870858      0
#> 478  -0.055537070      0
#> 479  -0.058807046      0
#> 480  -1.280751823      0
#> 481  -0.465841143      0
#> 482  -0.272488255      0
#> 483   0.133923960      0
#> 484   0.543951438      0
#> 485  -0.704170873      0
#> 486   0.796859380      0
#> 487   0.258773108      0
#> 488   2.266158933      0
#> 489  -0.598483683      0
#> 490   1.656204886      0
#> 491   2.149770552      0
#> 492   1.007116803      0
#> 493  -0.617626577      0
#> 494  -0.052551115      0
#> 495  -0.687446364      0
#> 496  -0.885936982      0
#> 497   0.028617402      0
#> 498   2.485231872      0
#> 499   0.267890682      0
#> 500   0.865649574      0
#> 501   1.030911719      0
#> 502  -0.296886495      0
#> 503  -1.004210510      0
#> 504  -0.424395536      0
#> 505  -0.492698820      0
#> 506   1.116852401      0
#> 507  -0.003430791      0
#> 508   0.056017411      0
#> 509   0.605268046      0
#> 510   1.320247777      0
#> 511   0.439702549      0
#> 512   0.035812966      0
#> 513   0.395325597      0
#> 514  -1.643731290      0
#> 515   0.858391547      0
#> 516  -0.543987885      0
#> 517  -0.704347551      0
#> 518   0.065491785      0
#> 519  -0.270717002      0
#> 520  -0.441630361      0
#> 521  -1.107666768      0
#> 522  -1.558103786      0
#> 523   0.245971676      0
#> 524  -0.943277638      0
#> 525   1.373544836      0
#> 526  -1.309064515      0
#> 527   0.560875409      0
#> 528  -0.097705458      0
#> 529  -0.154819040      0
#> 530  -0.466587983      0
#> 531  -1.474797218      0
#> 532  -1.481170932      0
#> 533   2.047311582      0
#> 534  -0.092762043      0
#> 535   0.339720211      0
#> 536  -0.118995276      0
#> 537  -0.862215932      0
#> 538  -0.170825773      0
#> 539  -0.156863678      0
#> 540   0.670209788      0
#> 541   0.402293348      0
#> 542   1.822938992      0
#> 543  -0.313537486      0
#> 544   1.264040330      0
#> 545   0.909302423      0
#> 546  -2.164022532      0
#> 547   0.505060861      0
#> 548   0.898238058      0
#> 549   0.212662537      0
#> 550  -0.212240828      0
#> 551  -1.031641595      0
#> 552   0.421666818      0
#> 553  -0.697315916      0
#> 554  -1.564972073      0
#> 555   0.446167418      0
#> 556  -0.108727417      0
#> 557  -0.970868605      0
#> 558   0.248274555      0
#> 559   1.026250685      0
#> 560  -0.374706697      0
#> 561  -0.670289162      0
#> 562  -0.276214773      0
#> 563  -0.789840851      0
#> 564   1.119382028      0
#> 565   0.552654476      0
#> 566   0.647435211      0
#> 567   0.068415589      0
#> 568  -0.291778930      0
#> 569   1.296797005      0
#> 570   0.908415564      0
#> 571   0.166843414      0
#> 572  -0.683018648      0
#> 573   0.636748429      0
#> 574   1.739454344      0
#> 575   0.268034323      0
#> 576  -1.457146052      0
#> 577  -0.091719780      0
#> 578  -0.178774491      0
#> 579   0.539487551      0
#> 580  -0.389832249      0
#> 581  -1.720608058      0
#> 582  -0.724663313      0
#> 583   1.331554674      0
#> 584   0.108956127      0
#> 585   0.624461989      0
#> 586  -0.036938390      0
#> 587   1.299530083      0
#> 588  -0.613977598      0
#> 589   0.232493721      0
#> 590   1.044735009      0
#> 591   1.656323442      0
#> 592  -0.062368781      0
#> 593   0.315936957      0
#> 594   1.626649905      0
#> 595   0.831958766      0
#> 596  -1.225602895      0
#> 597   0.381445355      0
#> 598   1.167516359      0
#> 599   0.369820191      0
#> 600   0.563186964      0
#> 601   0.701175475      0
#> 602   0.448461277      0
#> 603  -0.052482060      0
#> 604  -0.866192349      0
#> 605  -0.549281789      0
#> 606  -1.029515449      0
#> 607  -1.378204961      0
#> 608   0.489953983      0
#> 609  -0.444147572      0
#> 610  -1.581334608      0
#> 611   0.234046515      0
#> 612  -1.469112400      0
#> 613  -0.046686388      0
#> 614   0.505945379      0
#> 615   0.252741887      0
#> 616   0.671653443      0
#> 617   0.313001084      0
#> 618   0.038925642      0
#> 619   0.927719340      0
#> 620   0.819626667      0
#> 621   0.244869578      0
#> 622  -0.543899613      0
#> 623  -1.238116991      0
#> 624  -0.771695585      0
#> 625   0.035071533      0
#> 626   2.286000132      0
#> 627  -0.631938159      0
#> 628  -0.426263933      0
#> 629   0.242089528      0
#> 630  -0.483067803      0
#> 631  -0.181307183      0
#> 632   1.566797582      0
#> 633  -0.451887573      0
#> 634  -0.263373657      0
#> 635  -0.016489854      0
#> 636  -2.079379112      0
#> 637  -0.192460884      0
#> 638   0.938000384      0
#> 639   0.507858664      0
#> 640  -0.668028668      0
#> 641   1.908199424      0
#> 642  -0.101649374      0
#> 643  -0.070976831      0
#> 644   1.335883223      0
#> 645  -1.445981884      0
#> 646  -0.483696665      0
#> 647  -1.516225779      0
#> 648  -0.333610191      0
#> 649   1.007889713      0
#> 650  -0.349525558      0
#> 651   0.737947063      0
#> 652   0.189395846      0
#> 653  -0.362828349      0
#> 654  -0.843848294      0
#> 655   0.282794670      0
#> 656  -1.550957799      0
#> 657  -1.184898116      0
#> 658   0.098326378      0
#> 659  -0.393665725      0
#> 660  -2.852221195      0
#> 661  -0.028410253      0
#> 662  -0.599946885      0
#> 663   0.755709679      0
#> 664   0.335970144      0
#> 665  -0.481876958      0
#> 666  -0.811276529      0
#> 667   0.189961370      0
#> 668  -0.277299017      0
#> 669  -0.029876152      0
#> 670   0.085449837      0
#> 671   0.654639556      0
#> 672   1.641956031      0
#> 673  -0.826691940      0
#> 674   0.827987217      0
#> 675   0.733478198      0
#> 676   0.859434548      0
#> 677  -1.922296472      0
#> 678   0.146399636      0
#> 679  -0.355784065      0
#> 680  -0.028640911      0
#> 681  -1.430472610      0
#> 682   0.360477834      0
#> 683  -0.864182099      0
#> 684  -1.062049186      0
#> 685  -0.517525129      0
#> 686  -0.236414426      0
#> 687   0.816754822      0
#> 688  -1.019765191      0
#> 689   1.583309322      0
#> 690   0.644156566      0
#> 691   0.392766008      0
#> 692  -0.141949525      0
#> 693  -0.726318601      0
#> 694  -0.382891163      0
#> 695   2.396706913      0
#> 696   0.558058164      0
#> 697   0.901218604      0
#> 698  -0.297136869      0
#> 699  -0.541853852      0
#> 700   0.728929911      0
#> 701  -0.468467363      0
#> 702   0.772093874      0
#> 703  -0.234442571      0
#> 704  -1.684167608      0
#> 705  -1.744727221      0
#> 706  -0.751108800      0
#> 707  -0.171742578      0
#> 708  -1.874836518      0
#> 709   0.116726650      0
#> 710   0.251075039      0
#> 711  -2.659598070      0
#> 712   0.038455012      0
#> 713   0.148351084      0
#> 714  -1.726350663      0
#> 715  -0.116864822      0
#> 716  -0.031868308      0
#> 717   0.997496137      0
#> 718   1.178019163      0
#> 719   0.335413677      0
#> 720   0.470965085      0
#> 721  -0.062674490      0
#> 722  -0.071482122      0
#> 723  -0.822144992      0
#> 724  -2.284977829      0
#> 725  -0.172770975      0
#> 726  -0.072038221      0
#> 727  -0.364766146      0
#> 728   0.287932251      0
#> 729  -0.041491385      0
#> 730   1.230174661      0
#> 731   0.931404675      0
#> 732   0.957100715      0
#> 733   0.741273014      0
#> 734  -1.072281120      0
#> 735   0.960236836      0
#> 736  -0.567464301      0
#> 737  -0.800647660      0
#> 738   0.130912032      0
#> 739   0.730874588      0
#> 740  -0.252554007      0
#> 741  -0.129389019      0
#> 742   0.577902244      0
#> 743   0.018661555      0
#> 744   0.288383496      0
#> 745  -0.177852229      0
#> 746  -0.129295710      0
#> 747  -0.824454811      0
#> 748   0.599687195      0
#> 749   1.570696119      0
#> 750  -0.437870524      0
#> 751   0.625327321      0
#> 752  -0.910168795      0
#> 753  -2.212196648      0
#> 754   1.354670302      0
#> 755   0.254080196      0
#> 756  -0.271294203      0
#> 757   0.710093055      0
#> 758  -1.495197285      0
#> 759  -1.180257278      0
#> 760   1.399497334      0
#> 761  -1.139230408      0
#> 762  -1.384170745      0
#> 763  -0.004887881      0
#> 764   1.395837160      0
#> 765  -0.354614112      0
#> 766  -2.419094607      0
#> 767   0.311293347      0
#> 768  -1.087187987      0
#> 769   2.900883392      0
#> 770  -0.977120527      0
#> 771   0.878687107      0
#> 772   0.000859028      0
#> 773  -0.368696311      0
#> 774   0.358496393      0
#> 775   0.445035548      0
#> 776   1.165243329      0
#> 777  -0.476304581      0
#> 778  -2.441842206      0
#> 779  -1.582944705      0
#> 780  -0.144224049      0
#> 781   0.165175881      0
#> 782  -0.313825048      0
#> 783   0.182939780      0
#> 784   0.773902019      0
#> 785  -0.038806486      0
#> 786   1.324102468      0
#> 787   0.610153224      0
#> 788  -0.310908271      0
#> 789  -1.028520148      0
#> 790  -0.929548170      0
#> 791  -1.492230015      0
#> 792  -0.786167953      0
#> 793   0.689442217      0
#> 794   0.255293218      0
#> 795   1.043134285      0
#> 796  -0.242502612      0
#> 797   0.773510637      0
#> 798  -1.174292451      0
#> 799  -0.935379824      0
#> 800  -0.467805148      0
#> 801   0.805617574      0
#> 802   0.818977403      0
#> 803  -0.548065486      0
#> 804   0.268394697      0
#> 805   2.252273874      0
#> 806  -0.744408898      0
#> 807   1.329074921      0
#> 808   0.068612548      0
#> 809  -0.357065562      0
#> 810   0.817315363      0
#> 811   0.946673838      0
#> 812  -1.414042765      0
#> 813   0.399910315      0
#> 814  -0.178433276      0
#> 815  -1.936811960      0
#> 816   1.712741280      0
#> 817   0.972269865      0
#> 818   0.968817293      0
#> 819  -0.257275609      0
#> 820   1.172280832      0
#> 821   2.671795414      0
#> 822   0.205667474      0
#> 823  -1.368873950      0
#> 824  -1.954859477      0
#> 825  -0.405804304      0
#> 826  -0.112756344      0
#> 827   0.579091742      0
#> 828  -0.597355310      0
#> 829   1.151628197      0
#> 830   0.532490014      0
#> 831  -0.994746242      0
#> 832  -0.039618120      0
#> 833   0.245475233      0
#> 834  -0.355814593      0
#> 835   0.055003963      0
#> 836  -0.329839185      0
#> 837  -0.904736631      0
#> 838   0.929569254      0
#> 839   0.332754630      0
#> 840  -0.436827764      0
#> 841   0.261550651      0
#> 842  -0.699634203      0
#> 843   0.231422840      0
#> 844   1.063758511      0
#> 845  -0.334694463      0
#> 846   0.362794855      0
#> 847  -0.893645486      0
#> 848   0.825541811      0
#> 849   0.419698095      0
#> 850  -0.025921785      0
#> 851  -0.296450759      0
#> 852  -1.714304804      0
#> 853  -0.036585808      0
#> 854   0.758086238      0
#> 855  -0.169817179      0
#> 856   2.957659717      0
#> 857  -0.021309088      0
#> 858  -1.272633998      0
#> 859   0.739276588      0
#> 860  -1.129849835      0
#> 861   0.044546248      0
#> 862  -0.112081266      0
#> 863   0.777834473      0
#> 864  -1.019031265      0
#> 865  -0.888942267      0
#> 866  -2.194371004      0
#> 867   2.024414753      0
#> 868  -1.544392708      0
#> 869   0.849575205      0
#> 870  -1.029527652      0
#> 871  -0.153760831      0
#> 872  -0.293533992      0
#> 873   1.618790647      0
#> 874   0.302498112      0
#> 875  -0.469522797      0
#> 876   0.758432427      0
#> 877   1.204841288      0
#> 878   1.277552115      0
#> 879  -0.810362652      0
#> 880  -0.536796231      0
#> 881  -1.535853190      0
#> 882  -0.438469608      0
#> 883   1.868599770      0
#> 884   1.322966207      0
#> 885   0.535367192      0
#> 886   0.349480207      0
#> 887   0.089599798      0
#> 888  -0.901375011      0
#> 889   0.319495231      0
#> 890   0.651541157      0
#> 891   1.086509747      0
#> 892  -0.456399170      0
#> 893   1.165128070      0
#> 894   0.316360983      0
#> 895  -0.865045678      0
#> 896   0.558903626      0
#> 897   0.403420091      0
#> 898   0.573825058      0
#> 899   0.384532023      0
#> 900  -0.653433450      0
#> 901  -1.256144443      0
#> 902  -1.769395090      0
#> 903  -0.221312408      0
#> 904   0.916053617      0
#> 905  -0.777865171      0
#> 906  -1.982483627      0
#> 907   0.559562329      0
#> 908   0.895015299      0
#> 909   0.699741728      0
#> 910  -0.650448225      0
#> 911  -0.079625004      0
#> 912  -1.115799535      0
#> 913  -0.976027073      0
#> 914   0.946511830      0
#> 915   1.179979434      0
#> 916   1.142850367      0
#> 917   0.132516745      0
#> 918  -0.579829327      0
#> 919  -0.620594420      0
#> 920   1.000650315      0
#> 921   1.424752891      0
#> 922  -0.463732102      0
#> 923   0.129609183      0
#> 924   0.543917282      1
#> 925   1.115603511      1
#> 926   1.054163326      1
#> 927   0.908316819      1
#> 928   2.022720766      1
#> 929   0.080164745      1
#> 930  -0.489408451      1
#> 931   0.819363892      1
#> 932   1.379291889      1
#> 933  -0.538033919      1
#> 934   1.575494147      1
#> 935   0.503663693      1
#> 936   1.199263692      1
#> 937   1.974846707      1
#> 938   0.615807929      1
#> 939  -1.050577148      1
#> 940   0.024441938      1
#> 941  -0.651528737      1
#> 942   2.433876377      1
#> 943   1.258376732      1
#> 944   2.190169699      1
#> 945   0.402724689      1
#> 946   1.258194252      1
#> 947  -1.143598853      1
#> 948   0.623847706      1
#> 949  -1.708268677      1
#> 950  -1.302505204      1
#> 951   0.349912063      1
#> 952   2.169207335      1
#> 953  -0.553835692      1
#> 954  -0.374693992      1
#> 955   1.192097067      1
#> 956   1.571477171      1
#> 957   1.332003259      1
#> 958   1.669890300      1
#> 959   0.602673078      1
#> 960   1.770237804      1
#> 961   1.451550117      1
#> 962   1.722894060      1
#> 963   0.065280450      1
#> 964   3.368556579      1
#> 965   1.443704522      1
#> 966   0.464772350      1
#> 967   0.151661192      1
#> 968   1.155036658      1
#> 969   2.404296155      1
#> 970   1.680684794      1
#> 971  -0.245610343      1
#> 972   0.447199141      1
#> 973   1.170430992      1
#> 974   0.702676257      1
#> 975   0.860243441      1
#> 976   0.368080601      1
#> 977   0.450119109      1
#> 978   0.612314963      1
#> 979   1.476807390      1
#> 980   1.355547787      1
#> 981   0.767723257      1
#> 982   0.779818680      1
#> 983   3.142018403      1
#> 984  -1.104059412      1
#> 985   1.324250629      1
#> 986   0.332857732      1
#> 987   0.418955800      1
#> 988   0.582183225      1
#> 989   1.724331783      1
#> 990   1.698692160      1
#> 991   2.174803247      1
#> 992   0.858711449      1
#> 993  -1.394783274      1
#> 994   0.663635626      1
#> 995  -1.278961207      1
#> 996   1.628342093      1
#> 997   0.847718278      1
#> 998   2.226164682      1
#> 999   0.237129678      1
#> 1000 -0.840528770      1