Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.186247035      0
#> 2     0.379525960      0
#> 3     1.563999734      0
#> 4    -2.531693553      0
#> 5     0.053686438      0
#> 6     0.385710843      0
#> 7    -2.839585742      0
#> 8    -1.501147522      0
#> 9     0.105679166      0
#> 10    0.043507125      0
#> 11   -0.493021099      0
#> 12   -0.189225933      0
#> 13   -0.717700270      0
#> 14    0.537580285      0
#> 15    0.723791881      0
#> 16    0.833696908      0
#> 17    0.708201424      0
#> 18   -0.224180766      0
#> 19   -0.194918834      0
#> 20   -0.232419387      0
#> 21   -0.357874387      0
#> 22   -0.011181076      0
#> 23   -1.740468618      0
#> 24    1.147642283      0
#> 25    0.693302704      0
#> 26    0.937361334      0
#> 27   -0.643954229      0
#> 28    0.779960711      0
#> 29   -0.036917835      0
#> 30   -0.458159109      0
#> 31   -1.534042829      0
#> 32   -1.008054525      0
#> 33   -0.472556815      0
#> 34   -1.170236399      0
#> 35    0.907471735      0
#> 36    1.482760742      0
#> 37   -0.672305517      0
#> 38   -1.034908554      0
#> 39    0.355806857      0
#> 40   -0.123146829      0
#> 41   -0.646981931      0
#> 42    0.656478938      0
#> 43   -0.363900581      0
#> 44   -0.455691593      0
#> 45   -0.821762736      0
#> 46    0.915185001      0
#> 47    0.802234288      0
#> 48    1.271273768      0
#> 49    0.281163369      0
#> 50    0.838350170      0
#> 51   -1.172313689      0
#> 52    0.103802010      0
#> 53   -0.939840722      0
#> 54   -1.702834303      0
#> 55   -0.662493979      0
#> 56   -0.788724127      0
#> 57   -0.230591407      0
#> 58    1.231285312      0
#> 59   -1.790128747      0
#> 60    0.738383759      0
#> 61   -0.273527733      0
#> 62    0.139538029      0
#> 63    0.160378446      0
#> 64   -0.645497826      0
#> 65   -2.728135135      0
#> 66   -1.289628795      0
#> 67   -0.816718779      0
#> 68   -0.690686029      0
#> 69   -1.803662966      0
#> 70    0.509415340      0
#> 71   -0.152768110      0
#> 72    2.359340427      0
#> 73    0.766119767      0
#> 74   -0.537857379      0
#> 75   -0.082947405      0
#> 76    0.390649380      0
#> 77   -0.054878301      0
#> 78    0.107679341      0
#> 79    0.431898924      0
#> 80   -0.151150967      0
#> 81    0.328453323      0
#> 82   -0.020191916      0
#> 83   -1.254755689      0
#> 84    0.309109476      0
#> 85   -0.377073653      0
#> 86   -1.589326923      0
#> 87    0.414908923      0
#> 88   -0.131282999      0
#> 89   -0.408598566      0
#> 90   -1.228867846      0
#> 91    0.349173986      0
#> 92    1.018420027      0
#> 93   -0.217930211      0
#> 94    1.554639744      0
#> 95    2.387519516      0
#> 96    2.097355924      0
#> 97    1.139505321      0
#> 98   -0.359492828      0
#> 99   -0.170800535      0
#> 100  -1.941253645      0
#> 101   0.378917223      0
#> 102   1.453706265      0
#> 103   0.166744801      0
#> 104  -2.374075405      0
#> 105  -0.186512540      0
#> 106   0.866631332      0
#> 107   0.906881192      0
#> 108  -0.766485285      0
#> 109   0.382645792      0
#> 110  -0.668234749      0
#> 111   1.157847973      0
#> 112   0.242265316      0
#> 113  -0.134635576      0
#> 114  -0.493305397      0
#> 115   1.449451856      0
#> 116   0.602745748      0
#> 117   0.037014508      0
#> 118   0.117008512      0
#> 119   0.100194357      0
#> 120  -1.450176519      0
#> 121   0.779939403      0
#> 122  -0.226134551      0
#> 123   0.942488312      0
#> 124   3.221281993      0
#> 125  -0.274877666      0
#> 126  -0.460822524      0
#> 127   1.733581300      0
#> 128   0.530269684      0
#> 129   0.255248609      0
#> 130  -0.034048279      0
#> 131   1.850889492      0
#> 132   1.658332845      0
#> 133  -1.589637654      0
#> 134  -0.869562220      0
#> 135   0.988728843      0
#> 136  -1.421833717      0
#> 137  -1.251023188      0
#> 138  -1.417752031      0
#> 139  -0.153249072      0
#> 140   2.094529821      0
#> 141   2.175328615      0
#> 142   0.791302600      0
#> 143  -0.105516732      0
#> 144   1.107807621      0
#> 145  -0.900865757      0
#> 146   2.151546437      0
#> 147   2.451202450      0
#> 148  -0.258577588      0
#> 149  -0.631721810      0
#> 150  -0.053894154      0
#> 151   1.222263022      0
#> 152  -1.476952806      0
#> 153  -1.229799557      0
#> 154  -1.380454074      0
#> 155   0.123197950      0
#> 156  -1.242372466      0
#> 157  -0.718127287      0
#> 158  -1.784195255      0
#> 159  -0.514951466      0
#> 160   0.483564207      0
#> 161   1.982502299      0
#> 162   0.183283119      0
#> 163   1.320335055      0
#> 164  -0.467116917      0
#> 165  -0.343632515      0
#> 166   0.704241423      0
#> 167   0.365763386      0
#> 168   0.226466358      0
#> 169   0.558205083      0
#> 170   0.767943897      0
#> 171  -0.583810334      0
#> 172  -0.739883374      0
#> 173  -0.164454376      0
#> 174   0.693845535      0
#> 175   0.746497386      0
#> 176   1.896029832      0
#> 177   0.628887353      0
#> 178  -0.385366108      0
#> 179  -0.830611091      0
#> 180  -0.601912351      0
#> 181  -0.370031369      0
#> 182   1.538512970      0
#> 183   0.987201297      0
#> 184  -0.827035068      0
#> 185   0.667922278      0
#> 186  -0.135860659      0
#> 187  -0.962790687      0
#> 188  -0.600013140      0
#> 189   0.939079910      0
#> 190  -0.001980003      0
#> 191  -2.172607601      0
#> 192  -1.197890826      0
#> 193   1.289738961      0
#> 194   0.502117236      0
#> 195  -0.471458802      0
#> 196   0.601945772      0
#> 197   0.211122087      0
#> 198  -0.676520459      0
#> 199   1.488012974      0
#> 200   0.341955794      0
#> 201   0.944180024      0
#> 202   0.640837714      0
#> 203  -0.902633528      0
#> 204  -0.400686443      0
#> 205   2.068154931      0
#> 206  -1.890457456      0
#> 207  -0.734595070      0
#> 208   2.003508912      0
#> 209   1.833031341      0
#> 210   2.881883693      0
#> 211   0.694502689      0
#> 212   1.203198755      0
#> 213  -0.996865138      0
#> 214  -0.306028549      0
#> 215  -1.451446061      0
#> 216  -0.094194096      0
#> 217   0.846369912      0
#> 218  -0.812095377      0
#> 219   1.031469757      0
#> 220  -0.903235633      0
#> 221   0.055095127      0
#> 222  -0.319456488      0
#> 223   0.792706149      0
#> 224   0.216207386      0
#> 225   1.836686255      0
#> 226   2.722657036      0
#> 227   0.589615493      0
#> 228  -0.087902693      0
#> 229   2.183891171      0
#> 230  -0.321207393      0
#> 231   0.727744664      0
#> 232  -0.701608598      0
#> 233   0.196683253      0
#> 234  -1.401039459      0
#> 235  -0.018934408      0
#> 236   0.072211034      0
#> 237   0.370252573      0
#> 238  -1.002663152      0
#> 239   0.011097454      0
#> 240  -0.252100679      0
#> 241  -0.337174901      0
#> 242  -0.709762888      0
#> 243  -0.603744723      0
#> 244  -0.407301415      0
#> 245   1.357231135      0
#> 246   1.361565964      0
#> 247  -0.184973026      0
#> 248  -0.171509451      0
#> 249   0.831831952      0
#> 250  -2.058661054      0
#> 251  -1.759003160      0
#> 252  -0.756126297      0
#> 253  -0.718610245      0
#> 254   1.447781804      0
#> 255   0.783284184      0
#> 256   0.219130689      0
#> 257  -1.868649705      0
#> 258   0.251700910      0
#> 259   0.680579529      0
#> 260  -0.057684387      0
#> 261   0.869936902      0
#> 262  -1.270499212      0
#> 263  -0.452980922      0
#> 264  -1.359606503      0
#> 265   0.401195231      0
#> 266  -0.376496828      0
#> 267   1.519600146      0
#> 268   1.124788810      0
#> 269  -0.185601859      0
#> 270   0.093288943      0
#> 271   0.677368483      0
#> 272  -0.844826164      0
#> 273  -0.419603087      0
#> 274   0.554896389      0
#> 275   0.725078209      0
#> 276  -0.239427680      0
#> 277   1.237616416      0
#> 278  -0.812947259      0
#> 279   1.349923737      0
#> 280  -0.630537809      0
#> 281   0.040673631      0
#> 282   0.861742188      0
#> 283  -0.269747483      0
#> 284   1.171076609      0
#> 285  -1.670575265      0
#> 286   0.559515657      0
#> 287   0.221358226      0
#> 288   0.111494977      0
#> 289  -0.942533545      0
#> 290  -0.938547017      0
#> 291   2.290725131      0
#> 292   0.602975580      0
#> 293   0.934364727      0
#> 294   0.297896580      0
#> 295  -1.108531412      0
#> 296   1.729179543      0
#> 297  -0.949921479      0
#> 298   1.205239980      0
#> 299   0.853986857      0
#> 300  -1.148133147      0
#> 301  -1.770831712      0
#> 302  -0.253589135      0
#> 303  -0.035512602      0
#> 304  -0.055995920      0
#> 305  -0.153897120      0
#> 306   1.132659362      0
#> 307   0.559905952      0
#> 308  -0.060812102      0
#> 309   1.205689935      0
#> 310   1.115993005      0
#> 311   0.756404110      0
#> 312   0.021358468      0
#> 313   0.932439641      0
#> 314   0.160075305      0
#> 315  -0.165788485      0
#> 316   0.169851880      0
#> 317   0.547774308      0
#> 318   0.738909422      0
#> 319  -0.644489914      0
#> 320   0.091324801      0
#> 321   0.256526287      0
#> 322   0.320726843      0
#> 323   0.625590025      0
#> 324   1.131135384      0
#> 325  -1.783215243      0
#> 326   0.287828163      0
#> 327  -0.518940099      0
#> 328  -1.807442771      0
#> 329   0.155825053      0
#> 330  -0.034118292      0
#> 331  -0.375399298      0
#> 332   0.169100754      0
#> 333   1.713526387      0
#> 334   1.204363486      0
#> 335  -1.440681893      0
#> 336  -0.809216924      0
#> 337  -3.069183447      0
#> 338  -0.283395843      0
#> 339  -1.182689673      0
#> 340   1.684895130      0
#> 341  -0.387944808      0
#> 342   1.180933771      0
#> 343   0.748375948      0
#> 344  -0.467323776      0
#> 345  -1.181283716      0
#> 346   0.154733139      0
#> 347   1.367291066      0
#> 348  -1.799023907      0
#> 349  -0.584018086      0
#> 350   1.997121222      0
#> 351   0.486410804      0
#> 352  -1.293299316      0
#> 353   0.822135184      0
#> 354   1.041616158      0
#> 355   1.059007941      0
#> 356  -0.777609508      0
#> 357   0.590068776      0
#> 358  -0.603978672      0
#> 359  -0.391683135      0
#> 360  -1.360110473      0
#> 361  -0.430408506      0
#> 362   1.711368191      0
#> 363  -1.503513084      0
#> 364  -2.329225793      0
#> 365  -0.478936505      0
#> 366  -0.465970781      0
#> 367  -0.066578328      0
#> 368   0.794040792      0
#> 369  -0.656894338      0
#> 370   0.535213623      0
#> 371  -1.048072942      0
#> 372  -0.095071877      0
#> 373  -1.109971853      0
#> 374  -0.632535740      0
#> 375   1.233198524      0
#> 376  -0.237710148      0
#> 377   0.492848818      0
#> 378  -1.385427858      0
#> 379   1.260586872      0
#> 380   1.341072937      0
#> 381  -1.288639920      0
#> 382   0.775037673      0
#> 383  -0.434942804      0
#> 384   1.875289418      0
#> 385   0.121890292      0
#> 386  -1.272446567      0
#> 387  -1.397440721      0
#> 388   1.653664478      0
#> 389  -0.064061064      0
#> 390  -2.003017218      0
#> 391  -0.271708536      0
#> 392   1.436358567      0
#> 393  -0.398304882      0
#> 394  -1.555334734      0
#> 395  -2.345880234      0
#> 396   0.628341398      0
#> 397   0.740802687      0
#> 398   0.518804585      0
#> 399  -0.401386794      0
#> 400   2.956098069      0
#> 401  -1.659981298      0
#> 402  -0.508526834      0
#> 403  -0.636692687      0
#> 404   0.203899220      0
#> 405  -1.008553599      0
#> 406   1.180174111      0
#> 407  -0.380151905      0
#> 408   0.743089388      0
#> 409   0.677904074      0
#> 410   1.007249877      0
#> 411   1.631213804      0
#> 412   0.559314439      0
#> 413   0.958071537      0
#> 414  -1.060841078      0
#> 415  -1.916118327      0
#> 416   0.471359299      0
#> 417  -0.009252152      0
#> 418  -0.644555726      0
#> 419  -2.487733126      0
#> 420   2.014259900      0
#> 421   0.314321739      0
#> 422   0.568786899      0
#> 423   1.294536257      0
#> 424  -0.076913285      0
#> 425  -0.618909030      0
#> 426   0.190676749      0
#> 427  -1.507516608      0
#> 428  -2.176446239      0
#> 429  -0.516760652      0
#> 430  -1.957617441      0
#> 431   0.470867839      0
#> 432  -0.639810990      0
#> 433   0.540924291      0
#> 434   1.240433227      0
#> 435  -0.128623622      0
#> 436  -1.023886949      0
#> 437  -0.077256012      0
#> 438   1.009271550      0
#> 439  -0.305037402      0
#> 440  -1.969645366      0
#> 441   0.249685380      0
#> 442   0.249292971      0
#> 443   0.190453166      0
#> 444  -0.023130360      0
#> 445   0.442676980      0
#> 446  -0.400762756      0
#> 447   0.772312079      0
#> 448   0.409970494      0
#> 449   2.075092622      0
#> 450  -1.101976400      0
#> 451   0.302822540      0
#> 452   0.767648472      0
#> 453   1.841443713      0
#> 454  -0.633414526      0
#> 455   1.259306890      0
#> 456   1.110558445      0
#> 457  -1.533889577      0
#> 458  -0.239733504      0
#> 459  -0.280958415      0
#> 460   0.741978986      0
#> 461  -0.364599906      0
#> 462  -0.015012368      0
#> 463  -2.010602083      0
#> 464   0.361621743      0
#> 465   0.163753120      0
#> 466  -0.290785451      0
#> 467   0.952950677      0
#> 468  -0.108008656      0
#> 469  -0.375174759      0
#> 470   0.146055434      0
#> 471  -3.419119376      0
#> 472  -0.618110884      0
#> 473   0.683347676      0
#> 474  -0.250700333      0
#> 475   0.399667644      0
#> 476  -0.830320588      0
#> 477   0.319843167      0
#> 478  -1.455920756      0
#> 479  -1.199840625      0
#> 480   1.272452517      0
#> 481  -0.376925920      0
#> 482  -0.862485839      0
#> 483   0.379198692      0
#> 484  -0.639353760      0
#> 485  -0.611297866      0
#> 486  -0.546565522      0
#> 487  -0.079408025      0
#> 488   0.159763120      0
#> 489  -0.762485666      0
#> 490   1.723975575      0
#> 491  -0.352606193      0
#> 492   0.060135109      0
#> 493  -0.180202761      0
#> 494   0.019116531      0
#> 495   0.056362432      0
#> 496  -1.236220332      0
#> 497  -0.316904071      0
#> 498   0.720379343      0
#> 499   0.688617134      0
#> 500  -0.241323102      0
#> 501  -0.278628231      0
#> 502   1.225968936      0
#> 503   0.171488151      0
#> 504   0.611054058      0
#> 505   0.532742854      0
#> 506   0.219205862      0
#> 507  -1.318510132      0
#> 508  -0.062489257      0
#> 509  -0.874089807      0
#> 510   0.964565921      0
#> 511   0.698971906      0
#> 512  -0.844072678      0
#> 513   0.029107957      0
#> 514   0.053564393      0
#> 515  -0.429380997      0
#> 516  -0.232282820      0
#> 517   0.179255296      0
#> 518   0.095515576      0
#> 519   0.207322870      0
#> 520  -0.431835096      0
#> 521   1.872361586      0
#> 522  -0.487315048      0
#> 523  -0.896969365      0
#> 524  -1.062700982      0
#> 525   0.838462428      0
#> 526  -1.240573367      0
#> 527  -1.092471113      0
#> 528  -1.041554724      0
#> 529   0.161124862      0
#> 530  -1.357297007      0
#> 531  -0.583907373      0
#> 532  -0.209195755      0
#> 533   1.085856723      0
#> 534   1.767926957      0
#> 535   0.429252752      0
#> 536   0.445885408      0
#> 537  -0.216785462      0
#> 538  -1.773709144      0
#> 539  -0.208034509      0
#> 540  -0.668600935      0
#> 541   2.406093215      0
#> 542   0.165115357      0
#> 543   0.086113175      0
#> 544   1.807381115      0
#> 545  -0.802290009      0
#> 546   1.069945452      0
#> 547   1.700468751      0
#> 548  -0.731625832      0
#> 549   0.676475391      0
#> 550  -0.226729039      0
#> 551   1.661471240      0
#> 552  -0.157979464      0
#> 553  -0.317306106      0
#> 554  -1.198729957      0
#> 555  -0.732915658      0
#> 556  -0.107690367      0
#> 557  -0.665678987      0
#> 558  -0.193630120      0
#> 559   0.444020270      0
#> 560  -0.173881985      0
#> 561   1.131335104      0
#> 562  -0.087050065      0
#> 563   1.430672342      0
#> 564  -0.235451243      0
#> 565   1.128882503      0
#> 566   0.412614532      0
#> 567   1.296221089      0
#> 568   0.134284915      0
#> 569  -0.113035581      0
#> 570  -2.072647903      0
#> 571  -1.252370541      0
#> 572  -0.404696664      0
#> 573   0.187400917      0
#> 574   0.538700934      0
#> 575   0.754103375      0
#> 576  -0.508165177      0
#> 577   0.342033040      0
#> 578  -2.016502873      0
#> 579   1.380100099      0
#> 580   0.136169854      0
#> 581   1.295877167      0
#> 582   0.171820358      0
#> 583  -0.103766775      0
#> 584   0.011346783      0
#> 585  -1.573733897      0
#> 586   1.147037133      0
#> 587   0.825556274      0
#> 588   1.582733874      0
#> 589   2.580154255      0
#> 590   0.155264532      0
#> 591  -0.696315224      0
#> 592  -0.493448746      0
#> 593  -0.802956337      0
#> 594   0.621747527      0
#> 595   1.034370712      0
#> 596   0.282679527      0
#> 597  -1.329664809      0
#> 598  -1.187117301      0
#> 599   0.087991463      0
#> 600  -0.973835602      0
#> 601  -0.544142969      0
#> 602   0.759350870      0
#> 603   0.398349377      0
#> 604  -0.972009382      0
#> 605  -0.032977274      0
#> 606   0.778506029      0
#> 607  -1.045427535      0
#> 608  -0.626141945      0
#> 609  -0.069835671      0
#> 610   1.904453153      0
#> 611  -1.063893971      0
#> 612   0.564986914      0
#> 613  -0.208489879      0
#> 614   1.156423471      0
#> 615  -0.969789050      0
#> 616   1.006560545      0
#> 617   0.043263320      0
#> 618   0.481845427      0
#> 619  -0.705555117      0
#> 620   0.582718976      0
#> 621   0.276033581      0
#> 622   0.585293046      0
#> 623   0.334011898      0
#> 624   0.053513452      0
#> 625   1.225184000      0
#> 626   1.111433570      0
#> 627  -0.775276932      0
#> 628   1.213365785      0
#> 629   1.226078487      0
#> 630  -0.725744353      0
#> 631  -0.994348497      0
#> 632  -0.737496316      0
#> 633   0.484608395      0
#> 634  -3.562359088      0
#> 635   0.218422040      0
#> 636   0.428236944      0
#> 637  -0.382848901      0
#> 638   0.224324644      0
#> 639   1.194622697      0
#> 640  -0.200535271      0
#> 641   0.953042117      0
#> 642   0.905606299      0
#> 643   2.409485958      0
#> 644   1.043167752      0
#> 645  -0.650996843      0
#> 646   2.126955610      0
#> 647  -1.751891455      0
#> 648  -0.401099405      0
#> 649  -0.852265371      0
#> 650   0.779791718      0
#> 651   0.216956062      0
#> 652  -0.332979396      0
#> 653   0.335307055      0
#> 654  -0.616130184      0
#> 655   1.131696766      0
#> 656  -0.324004661      0
#> 657  -1.919030722      0
#> 658  -0.978708794      0
#> 659   0.071161464      0
#> 660   0.179820154      0
#> 661   0.411880460      0
#> 662   0.299446601      0
#> 663   1.436075981      0
#> 664  -0.432488509      0
#> 665   0.994120407      0
#> 666   0.904768347      0
#> 667   0.411181967      0
#> 668   0.774613046      0
#> 669   0.486801568      0
#> 670   0.312577298      0
#> 671   0.255742294      0
#> 672  -0.269140237      0
#> 673  -0.673338568      0
#> 674   0.455189808      0
#> 675   1.322753488      0
#> 676  -0.046718765      0
#> 677   0.875256191      0
#> 678   1.241210458      0
#> 679   0.404808605      0
#> 680  -0.356534180      0
#> 681   1.348263055      0
#> 682   1.577213016      0
#> 683   1.029561372      0
#> 684  -1.590253603      0
#> 685   0.030588121      0
#> 686  -0.423084280      0
#> 687   1.360055549      0
#> 688   1.628709533      0
#> 689   2.091721556      0
#> 690   0.534383794      0
#> 691  -0.060918213      0
#> 692  -0.796155661      0
#> 693  -1.295334282      0
#> 694   0.971684980      0
#> 695  -1.320260084      0
#> 696  -0.966964317      0
#> 697  -0.577778244      0
#> 698  -0.825192958      0
#> 699   1.183010256      0
#> 700  -0.489318198      0
#> 701  -0.204980405      0
#> 702   0.564025212      0
#> 703  -0.007215686      0
#> 704  -0.721712204      0
#> 705  -0.688923275      0
#> 706   0.241660772      0
#> 707  -1.018869503      0
#> 708   0.925712929      0
#> 709  -0.520363948      0
#> 710  -0.090439650      0
#> 711   1.232160031      0
#> 712  -0.499817514      0
#> 713   0.936135261      0
#> 714   0.577537687      0
#> 715   1.073882134      0
#> 716  -1.961495818      0
#> 717   0.480957825      0
#> 718  -1.665352422      0
#> 719  -0.521647737      0
#> 720  -0.800784140      0
#> 721   2.527091084      0
#> 722   0.573104634      0
#> 723   1.344525333      0
#> 724  -0.858979576      0
#> 725  -0.170464725      0
#> 726   0.745412683      0
#> 727  -0.066874228      0
#> 728  -1.321936966      0
#> 729  -0.616000168      0
#> 730   1.165297251      0
#> 731  -0.110947121      0
#> 732  -0.201033742      0
#> 733   1.018270964      0
#> 734   1.404213431      0
#> 735  -0.858128796      0
#> 736  -0.383306567      0
#> 737   0.198389485      0
#> 738   1.968854873      0
#> 739  -0.786724788      0
#> 740  -0.159474165      0
#> 741  -1.308503614      0
#> 742  -0.860784664      0
#> 743   0.146919578      0
#> 744  -0.233146549      0
#> 745   0.800008654      0
#> 746   0.607991952      0
#> 747  -1.064218673      0
#> 748   1.038849274      0
#> 749   1.162432642      0
#> 750  -0.971697557      0
#> 751   0.796243333      0
#> 752   1.240842683      0
#> 753  -0.538746007      0
#> 754  -1.122750262      0
#> 755   0.130976804      0
#> 756   0.761871131      0
#> 757   0.737715099      0
#> 758  -1.376369049      0
#> 759   0.410829834      0
#> 760  -1.993190349      0
#> 761  -0.540666066      0
#> 762  -0.674916266      0
#> 763   0.633958493      0
#> 764   0.238104799      0
#> 765   0.646501877      0
#> 766   1.015163174      0
#> 767  -0.969646330      0
#> 768   0.262396946      0
#> 769   2.227793095      0
#> 770   1.254928842      0
#> 771  -0.381390936      0
#> 772   0.560049308      0
#> 773   0.390106169      0
#> 774   0.576460383      0
#> 775   0.638847412      0
#> 776   1.300946721      0
#> 777   0.379496406      0
#> 778   0.044565991      0
#> 779  -0.252777624      0
#> 780  -0.829240037      0
#> 781   1.342883917      0
#> 782  -2.346636748      0
#> 783   0.433587158      0
#> 784   1.717644690      0
#> 785  -1.724646434      0
#> 786   0.096738693      0
#> 787  -0.856085269      0
#> 788   1.065777610      0
#> 789   0.294307138      0
#> 790  -0.114342616      0
#> 791  -0.262679153      0
#> 792  -0.507851915      0
#> 793   2.164469500      0
#> 794  -0.421714885      0
#> 795   0.088089544      0
#> 796   0.040981307      0
#> 797  -0.281122474      0
#> 798   1.754013385      0
#> 799  -1.102879066      0
#> 800  -1.202986139      0
#> 801   0.134097797      0
#> 802  -0.824307495      0
#> 803  -0.271723368      0
#> 804   0.328184146      0
#> 805   0.958655865      0
#> 806  -0.480744023      0
#> 807   0.331047522      0
#> 808  -0.383318226      0
#> 809   0.615498238      0
#> 810   1.058708688      0
#> 811   0.138283573      0
#> 812  -1.099885413      0
#> 813   0.073433537      0
#> 814   0.759287917      0
#> 815   1.937518855      0
#> 816   0.910383576      0
#> 817  -0.947431197      0
#> 818   0.703060939      0
#> 819  -0.716483050      0
#> 820   0.859904242      0
#> 821   0.552033371      0
#> 822   1.269014795      0
#> 823   0.956284230      0
#> 824  -0.356490123      0
#> 825  -0.632241895      0
#> 826  -0.406490202      0
#> 827  -0.772570119      0
#> 828  -1.903529803      0
#> 829   0.957794490      0
#> 830   0.007956025      0
#> 831  -1.531276658      0
#> 832   2.051078957      0
#> 833   0.877752207      0
#> 834  -1.033147666      0
#> 835  -0.799609049      0
#> 836   1.640732426      0
#> 837   0.226843367      0
#> 838  -0.549532418      0
#> 839   1.275307983      0
#> 840  -0.870430320      0
#> 841   0.904471872      0
#> 842   1.540127914      0
#> 843   0.833097623      0
#> 844  -2.687981083      0
#> 845   0.118523763      0
#> 846  -1.004073782      0
#> 847  -0.293023145      0
#> 848   0.304682086      0
#> 849   1.714243650      0
#> 850   0.671257275      0
#> 851   1.643171988      0
#> 852   0.558640810      0
#> 853   2.715501954      0
#> 854  -2.002034943      0
#> 855  -0.718684320      0
#> 856  -1.544396301      0
#> 857   0.135349685      0
#> 858  -0.544024637      0
#> 859   1.236219525      0
#> 860  -1.236917347      0
#> 861   0.514682933      0
#> 862  -0.701608505      0
#> 863   1.087927324      0
#> 864   0.734788190      0
#> 865   0.522307457      0
#> 866   0.903754067      0
#> 867   0.442966391      0
#> 868   1.335834446      0
#> 869   2.036388003      0
#> 870  -1.181546922      0
#> 871  -1.124028896      0
#> 872   0.750842923      0
#> 873   1.685999189      0
#> 874   2.051137717      0
#> 875   0.977038204      0
#> 876   0.208715956      0
#> 877   1.394430659      0
#> 878   0.170634584      0
#> 879   1.278668449      0
#> 880   0.343191751      0
#> 881   0.508992220      0
#> 882  -0.151198657      0
#> 883   0.303587336      0
#> 884  -1.260153601      0
#> 885   0.765701468      0
#> 886   1.271307039      0
#> 887   0.740758394      0
#> 888  -0.330743887      0
#> 889  -0.643127368      0
#> 890   1.028285911      0
#> 891  -0.423307925      0
#> 892  -0.026656196      1
#> 893   2.121148278      1
#> 894   1.779390027      1
#> 895  -0.568911156      1
#> 896   1.805411884      1
#> 897   0.952104001      1
#> 898   0.713801098      1
#> 899  -0.609675851      1
#> 900  -0.026098201      1
#> 901   2.217787029      1
#> 902  -0.462138094      1
#> 903   0.024259631      1
#> 904   0.535403101      1
#> 905   1.561179486      1
#> 906   1.036599208      1
#> 907  -0.963966425      1
#> 908  -1.333940911      1
#> 909   1.657345377      1
#> 910   0.235573537      1
#> 911   1.241583037      1
#> 912   0.738386341      1
#> 913   3.084496247      1
#> 914  -1.022875298      1
#> 915  -0.339382835      1
#> 916   2.790857517      1
#> 917  -0.252547116      1
#> 918  -0.656922386      1
#> 919   2.005225737      1
#> 920   1.153026625      1
#> 921   1.565601859      1
#> 922  -0.940037780      1
#> 923   1.198433496      1
#> 924  -0.634504248      1
#> 925   1.400583173      1
#> 926   1.805257867      1
#> 927   1.626884800      1
#> 928   1.553328962      1
#> 929   0.828273273      1
#> 930  -0.208393057      1
#> 931  -0.584142085      1
#> 932   1.166211741      1
#> 933   0.982796883      1
#> 934  -1.678598450      1
#> 935   1.784718835      1
#> 936  -0.232336815      1
#> 937   1.061400400      1
#> 938  -0.797180829      1
#> 939   0.535691913      1
#> 940   1.942274477      1
#> 941  -0.979878841      1
#> 942  -0.094684437      1
#> 943   0.438105162      1
#> 944   1.683441898      1
#> 945   0.949957088      1
#> 946   2.011593276      1
#> 947   0.425511312      1
#> 948   0.225191373      1
#> 949  -0.263967798      1
#> 950   0.894420147      1
#> 951  -0.871114066      1
#> 952  -0.356941993      1
#> 953   1.498643367      1
#> 954   1.668634610      1
#> 955   0.575427658      1
#> 956   1.400091232      1
#> 957   0.956693085      1
#> 958  -0.287602518      1
#> 959   1.484118868      1
#> 960   1.198985894      1
#> 961   0.162041258      1
#> 962  -0.566619223      1
#> 963  -0.108402674      1
#> 964   0.741532517      1
#> 965  -0.337414669      1
#> 966   0.084158507      1
#> 967   1.282858332      1
#> 968  -0.368167565      1
#> 969   0.887575151      1
#> 970   1.116461269      1
#> 971   0.841788845      1
#> 972   1.285601921      1
#> 973   2.136883803      1
#> 974  -0.528335318      1
#> 975   0.364648587      1
#> 976   1.534115680      1
#> 977   1.370231900      1
#> 978   0.955867243      1
#> 979   2.419701147      1
#> 980  -2.217773120      1
#> 981   2.816951724      1
#> 982  -0.584399684      1
#> 983   0.233658037      1
#> 984   1.872280798      1
#> 985   3.547663309      1
#> 986  -0.250654004      1
#> 987   1.420264733      1
#> 988   0.505808058      1
#> 989   0.337503306      1
#> 990   1.335582134      1
#> 991   0.372767327      1
#> 992   1.579717233      1
#> 993   0.865442541      1
#> 994   2.378371986      1
#> 995   1.706402885      1
#> 996   0.008456925      1
#> 997  -0.389446784      1
#> 998  -0.226267292      1
#> 999  -0.053574889      1
#> 1000  0.068548851      1