Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     0.7030646563      0
#> 2     0.3441674748      0
#> 3    -0.3470254315      0
#> 4    -2.1281514352      0
#> 5    -0.2109429630      0
#> 6    -1.7442772996      0
#> 7    -0.0156451658      0
#> 8    -0.3837406189      0
#> 9    -1.4071047311      0
#> 10   -0.1917866878      0
#> 11   -0.0710057528      0
#> 12    0.4317257034      0
#> 13   -0.5278690518      0
#> 14   -1.3137749507      0
#> 15   -1.4007304106      0
#> 16    0.3405872547      0
#> 17   -0.0869075603      0
#> 18   -2.0117542335      0
#> 19    2.4487352104      0
#> 20   -1.0058308657      0
#> 21   -1.4205999712      0
#> 22   -0.1135190581      0
#> 23   -0.7623044490      0
#> 24    0.3177888396      0
#> 25   -1.1785249790      0
#> 26    0.6676055307      0
#> 27   -0.0361368101      0
#> 28   -1.5459719752      0
#> 29    1.3948119055      0
#> 30    2.0919558087      0
#> 31    0.2990008398      0
#> 32    1.1075974249      0
#> 33    0.5650561292      0
#> 34   -0.4248797706      0
#> 35   -0.5408026296      0
#> 36    0.5240801355      0
#> 37   -1.0141516080      0
#> 38   -0.3838234386      0
#> 39   -0.8742224223      0
#> 40    0.3030390369      0
#> 41    0.2514890705      0
#> 42   -0.1274216782      0
#> 43   -1.0461341176      0
#> 44   -0.4683726902      0
#> 45    0.5059832211      0
#> 46   -1.2561916767      0
#> 47   -0.2493831931      0
#> 48   -0.0651726156      0
#> 49    1.1060181895      0
#> 50   -0.0751435862      0
#> 51   -1.1027308111      0
#> 52    0.1801564639      0
#> 53   -1.6174648462      0
#> 54   -0.4038323422      0
#> 55   -1.6287559538      0
#> 56   -0.8722253431      0
#> 57   -1.1946036875      0
#> 58   -0.5634874722      0
#> 59    0.3812163913      0
#> 60    0.4006540444      0
#> 61    0.6435660642      0
#> 62    3.4015454751      0
#> 63   -1.6638103730      0
#> 64   -0.3301336148      0
#> 65   -0.4307523699      0
#> 66   -0.4773566792      0
#> 67   -0.0009968714      0
#> 68   -0.6827716758      0
#> 69   -0.0203155787      0
#> 70    1.3087366750      0
#> 71   -1.9407859959      0
#> 72   -2.6365095338      0
#> 73    0.3169097643      0
#> 74    0.4798217433      0
#> 75    1.4708371641      0
#> 76    0.6102058842      0
#> 77    1.2491115829      0
#> 78    0.5351954441      0
#> 79    1.1956548727      0
#> 80    0.3146430630      0
#> 81   -2.3676271167      0
#> 82    1.6210621072      0
#> 83    0.4173060026      0
#> 84    0.5060730841      0
#> 85    0.6070685714      0
#> 86   -0.2672812006      0
#> 87    0.7582654996      0
#> 88    0.9841693558      0
#> 89    1.8877396633      0
#> 90    2.1920864546      0
#> 91    0.3209107566      0
#> 92   -2.0952175984      0
#> 93   -0.3606327536      0
#> 94    0.3084024748      0
#> 95   -0.3651034464      0
#> 96   -0.2378908831      0
#> 97    0.3347903213      0
#> 98   -0.2244467915      0
#> 99   -1.0974203818      0
#> 100  -0.7761655997      0
#> 101   1.7965408533      0
#> 102  -0.8682382355      0
#> 103   0.4199145806      0
#> 104  -0.1882946529      0
#> 105  -0.1117805817      0
#> 106  -0.2717480850      0
#> 107   0.1136445661      0
#> 108  -0.1598847800      0
#> 109   1.2190293666      0
#> 110  -0.0839135979      0
#> 111  -1.7316475659      0
#> 112   0.6659289870      0
#> 113  -0.1048608108      0
#> 114  -0.1765861900      0
#> 115   1.9524349360      0
#> 116  -0.1171260598      0
#> 117   1.8584069813      0
#> 118  -2.0351466691      0
#> 119   1.4289670322      0
#> 120   0.0264211761      0
#> 121  -0.8580989545      0
#> 122  -0.7704856138      0
#> 123  -1.1269360708      0
#> 124  -0.3458888196      0
#> 125  -0.0994678348      0
#> 126   0.3618732742      0
#> 127  -0.6958393934      0
#> 128  -0.6115232938      0
#> 129  -0.0370766830      0
#> 130  -0.0369426692      0
#> 131   1.1169673486      0
#> 132  -0.2679532560      0
#> 133   0.2931275765      0
#> 134  -1.2712966779      0
#> 135  -0.3347417309      0
#> 136  -1.5076580394      0
#> 137   0.0032155714      0
#> 138   0.8359492753      0
#> 139   1.4689585261      0
#> 140   0.0564512992      0
#> 141   0.9655198713      0
#> 142  -1.0191982168      0
#> 143   0.8094959484      0
#> 144   0.0988405029      0
#> 145   0.4175218507      0
#> 146  -0.4466714554      0
#> 147  -1.0535303206      0
#> 148  -1.6841465737      0
#> 149  -1.3289535998      0
#> 150   1.6043503110      0
#> 151  -0.8221840995      0
#> 152   0.7098921242      0
#> 153  -1.6843025899      0
#> 154   1.0271501091      0
#> 155   3.0088980716      0
#> 156  -0.1556235241      0
#> 157  -0.1874554342      0
#> 158   0.2383571365      0
#> 159   1.1518096530      0
#> 160   1.0835155004      0
#> 161  -0.7443197240      0
#> 162   1.8789802569      0
#> 163  -1.0671900046      0
#> 164  -1.3670243118      0
#> 165  -0.9819508141      0
#> 166  -1.2331318997      0
#> 167   0.0656118064      0
#> 168  -0.2441504774      0
#> 169   0.2679145055      0
#> 170  -0.9577545274      0
#> 171  -0.4528032308      0
#> 172  -0.5084633938      0
#> 173   2.0465352475      0
#> 174   0.7632690784      0
#> 175  -0.8329600756      0
#> 176   0.5914874241      0
#> 177   1.5163538361      0
#> 178   0.3561427710      0
#> 179  -0.3480565052      0
#> 180   1.1194356122      0
#> 181   0.4298124179      0
#> 182  -0.3330095050      0
#> 183  -0.8554255184      0
#> 184  -0.4489472425      0
#> 185  -0.2552370618      0
#> 186  -1.6124316094      0
#> 187   0.7677660167      0
#> 188  -0.5533083825      0
#> 189   0.0763793120      0
#> 190  -0.7657233882      0
#> 191  -0.9186289711      0
#> 192  -1.0340261856      0
#> 193  -0.2647860976      0
#> 194   0.9325703606      0
#> 195  -0.6130918770      0
#> 196   1.5662967019      0
#> 197  -0.2547848834      0
#> 198  -0.4495732996      0
#> 199  -2.1251267004      0
#> 200  -0.2171098259      0
#> 201   1.0582619418      0
#> 202  -1.1871076288      0
#> 203  -1.2536843739      0
#> 204   0.7791493299      0
#> 205   1.2656664503      0
#> 206   1.5563003505      0
#> 207   0.6556435852      0
#> 208  -1.0209013302      0
#> 209   0.2198853202      0
#> 210  -0.5557596504      0
#> 211   0.8642815701      0
#> 212  -0.0259713844      0
#> 213   0.3534300333      0
#> 214   0.7951186800      0
#> 215  -0.7146362663      0
#> 216   0.4390005868      0
#> 217  -0.9484911282      0
#> 218  -0.3799147949      0
#> 219  -0.2115724404      0
#> 220   0.2832471958      0
#> 221   0.4631826687      0
#> 222   1.4770913625      0
#> 223  -0.0921458151      0
#> 224   0.8306375683      0
#> 225   1.3675797175      0
#> 226  -0.0555551434      0
#> 227  -0.4401404599      0
#> 228   0.4588458310      0
#> 229  -0.4539604963      0
#> 230   1.5687464909      0
#> 231   0.5979458290      0
#> 232   0.9019230515      0
#> 233  -1.3143387772      0
#> 234  -1.2191811460      0
#> 235  -1.0096100858      0
#> 236   0.3890029877      0
#> 237   1.4981030493      0
#> 238  -0.7186059534      0
#> 239  -2.1747775863      0
#> 240  -0.9525091354      0
#> 241  -0.3399116335      0
#> 242   1.7399983005      0
#> 243   0.0345632199      0
#> 244  -0.5905269973      0
#> 245   0.6486510611      0
#> 246  -2.2188635147      0
#> 247  -0.2050399911      0
#> 248   0.0344048958      0
#> 249   0.8944177728      0
#> 250  -2.1527699885      0
#> 251   0.3550573545      0
#> 252   2.9106493667      0
#> 253  -1.0058357296      0
#> 254  -0.4352431314      0
#> 255   1.8728301861      0
#> 256   0.4093309000      0
#> 257  -0.2743704178      0
#> 258  -1.7045911164      0
#> 259  -0.1277365121      0
#> 260   0.0299810901      0
#> 261  -1.9879192626      0
#> 262  -0.5141779949      0
#> 263  -1.8211106770      0
#> 264   1.9772407370      0
#> 265  -0.4930489788      0
#> 266   1.3254350973      0
#> 267   0.5862292511      0
#> 268  -2.2257277106      0
#> 269   0.5129554400      0
#> 270   0.3163147411      0
#> 271   1.0981186061      0
#> 272   0.3148411050      0
#> 273   0.1296694594      0
#> 274  -0.5142816845      0
#> 275   1.4838208479      0
#> 276  -0.2244198142      0
#> 277  -0.5622566525      0
#> 278  -1.3428770412      0
#> 279  -0.4939647245      0
#> 280   0.0349498683      0
#> 281  -1.9278241978      0
#> 282   0.3070047411      0
#> 283   1.7848049781      0
#> 284   0.7450458979      0
#> 285  -1.6122948240      0
#> 286  -0.7017075075      0
#> 287  -2.3353713151      0
#> 288   0.1622064594      0
#> 289   1.0908835740      0
#> 290  -1.5447432749      0
#> 291   0.1132848817      0
#> 292  -0.6186387708      0
#> 293   0.0724733765      0
#> 294  -0.7376153024      0
#> 295  -1.2335809536      0
#> 296  -1.4817086041      0
#> 297   0.5040294743      0
#> 298   1.3537674664      0
#> 299   1.0981453693      0
#> 300   0.4122624858      0
#> 301  -0.8616003451      0
#> 302  -0.9486168267      0
#> 303  -0.5733608347      0
#> 304   0.3475216392      0
#> 305  -0.5617241540      0
#> 306   0.4189745935      0
#> 307   0.3917259486      0
#> 308  -0.6339373651      0
#> 309   0.8129554024      0
#> 310  -1.3066851735      0
#> 311   1.8825825972      0
#> 312  -0.6896856430      0
#> 313   1.2174661426      0
#> 314   1.9986418841      0
#> 315  -0.0825386835      0
#> 316  -2.1806266739      0
#> 317  -0.6749990039      0
#> 318  -0.6234315135      0
#> 319   2.1103452952      0
#> 320  -1.1087151316      0
#> 321   1.4234159370      0
#> 322   1.4963017095      0
#> 323  -0.8353387401      0
#> 324  -1.3820449223      0
#> 325   2.9772290379      0
#> 326  -0.5677395255      0
#> 327   2.0398630395      0
#> 328   0.3747346046      0
#> 329   3.5264905650      0
#> 330  -0.0086293771      0
#> 331   1.1000021261      0
#> 332   1.3787346314      0
#> 333   1.4310803395      0
#> 334   0.8300524811      0
#> 335  -1.4572246871      0
#> 336   0.8006534198      0
#> 337   1.0496794272      0
#> 338  -1.0942428693      0
#> 339  -0.2946314203      0
#> 340   0.6338042193      0
#> 341  -0.3902276945      0
#> 342   1.8699142457      0
#> 343   1.1784159984      0
#> 344  -0.2304578717      0
#> 345  -0.3856144638      0
#> 346  -1.4926379013      0
#> 347  -0.8172211624      0
#> 348  -0.0400684901      0
#> 349   0.0352920300      0
#> 350   0.6454217941      0
#> 351  -0.5736531948      0
#> 352  -2.3868973571      0
#> 353  -0.8199651465      0
#> 354  -0.4378126052      0
#> 355  -0.2877219152      0
#> 356   0.4184561743      0
#> 357  -0.8255295852      0
#> 358  -1.4596003544      0
#> 359  -2.1047725786      0
#> 360  -0.0635410123      0
#> 361  -0.0361539125      0
#> 362   0.6650759284      0
#> 363  -0.9270459146      0
#> 364   1.0192185438      0
#> 365   0.9221615795      0
#> 366  -0.0468288495      0
#> 367   0.2178784252      0
#> 368  -0.3087704378      0
#> 369   0.1042449707      0
#> 370   0.5501525985      0
#> 371   0.3332914816      0
#> 372   1.1806065519      0
#> 373   0.2512527657      0
#> 374   0.6170354087      0
#> 375   0.4157861085      0
#> 376   0.1378982947      0
#> 377  -1.1285611452      0
#> 378   0.3631751017      0
#> 379  -1.4554757303      0
#> 380   1.2118059243      0
#> 381   0.7621011547      0
#> 382   0.5108673789      0
#> 383  -0.2065490289      0
#> 384  -1.6008279642      0
#> 385   0.0366723092      0
#> 386  -0.1806168783      0
#> 387  -2.7945817623      0
#> 388   0.3421860387      0
#> 389   0.5455657336      0
#> 390   0.2341083414      0
#> 391  -0.1961749792      0
#> 392  -0.9363921444      0
#> 393  -0.8104123126      0
#> 394  -0.2255684285      0
#> 395  -0.1554562808      0
#> 396   0.8730812080      0
#> 397  -0.7283842243      0
#> 398  -0.5981961743      0
#> 399   0.5857652226      0
#> 400   0.2020840638      0
#> 401   0.9024202779      0
#> 402   0.5857389699      0
#> 403   1.1041592362      0
#> 404  -0.4887915922      0
#> 405  -0.3301146622      0
#> 406   0.4155810023      0
#> 407   0.1866859392      0
#> 408   0.4339678525      0
#> 409   0.3730061216      0
#> 410   0.0645288061      0
#> 411  -0.9816191044      0
#> 412   0.5216524045      0
#> 413  -0.6128788199      0
#> 414  -0.2330526622      0
#> 415   0.1082562541      0
#> 416  -1.1306490837      0
#> 417   0.0242000767      0
#> 418  -0.1597475226      0
#> 419  -0.8707337495      0
#> 420   0.6939216099      0
#> 421  -0.0174653629      0
#> 422  -0.1201925781      0
#> 423   0.9255186375      0
#> 424  -0.0653866528      0
#> 425  -0.0925024764      0
#> 426   0.3887800766      0
#> 427   0.4126205326      0
#> 428   0.4012864867      0
#> 429   0.4801178556      0
#> 430  -0.5724093180      0
#> 431   1.0408101063      0
#> 432   0.4104036980      0
#> 433  -0.9406668313      0
#> 434  -0.5980133498      0
#> 435  -0.1149126586      0
#> 436  -0.2363400919      0
#> 437   1.1548101926      0
#> 438  -0.8821503113      0
#> 439   0.7376693166      0
#> 440   0.0435804195      0
#> 441   0.0579259353      0
#> 442   0.1195801233      0
#> 443   0.2940774165      0
#> 444  -0.5163438262      0
#> 445  -2.7164614567      0
#> 446   1.6445923435      0
#> 447  -0.2554269807      0
#> 448  -0.1759730992      0
#> 449   0.5596980254      0
#> 450  -0.8051067675      0
#> 451  -0.1442867985      0
#> 452   0.1620152698      0
#> 453   0.5140869774      0
#> 454  -0.6534007386      0
#> 455   0.6638948426      0
#> 456   0.8603789976      0
#> 457   0.4846954988      0
#> 458   1.4852693566      0
#> 459   0.2134884359      0
#> 460  -0.3286387806      0
#> 461  -0.2037992560      0
#> 462  -0.7900060571      0
#> 463   1.8919134141      0
#> 464  -0.4095904444      0
#> 465   0.6994629787      0
#> 466   0.0415566432      0
#> 467  -0.3070448453      0
#> 468   0.7323618138      0
#> 469  -1.2750993294      0
#> 470   0.8263704889      0
#> 471   1.0835101441      0
#> 472  -0.8465036387      0
#> 473  -0.6232249023      0
#> 474   0.0212166658      0
#> 475   1.9794482656      0
#> 476  -0.9782478716      0
#> 477  -0.6028928034      0
#> 478   3.3617837211      0
#> 479  -0.8468611229      0
#> 480  -0.2242036992      0
#> 481   0.5424485859      0
#> 482  -0.9123761565      0
#> 483   1.5490017473      0
#> 484   0.0963176110      0
#> 485   0.1763906139      0
#> 486  -0.2280823767      0
#> 487   0.2625130528      0
#> 488  -1.0642787865      0
#> 489  -0.6735156634      0
#> 490   2.1822209657      0
#> 491  -1.5996015833      0
#> 492   0.2100223928      0
#> 493  -1.7151401834      0
#> 494  -1.5178773636      0
#> 495  -0.0664722141      0
#> 496  -2.4634889479      0
#> 497   0.5637890307      0
#> 498  -0.3148815559      0
#> 499   0.1665803599      0
#> 500  -0.3781412137      0
#> 501  -0.0895815516      0
#> 502  -0.3393769380      0
#> 503  -0.2515684534      0
#> 504  -2.3611478561      0
#> 505   2.6178658464      0
#> 506   0.2373307255      0
#> 507  -1.7461094941      0
#> 508   0.1539325819      0
#> 509   0.9142273411      0
#> 510  -0.7627498832      0
#> 511  -0.3009019366      0
#> 512   0.0493197406      0
#> 513  -0.5588016912      0
#> 514  -0.7710558334      0
#> 515   0.0095352935      0
#> 516  -0.5952372550      0
#> 517   1.3045739925      0
#> 518  -1.6107009552      0
#> 519  -0.3354233392      0
#> 520  -1.9366158299      0
#> 521   1.6121585814      0
#> 522  -0.4768324122      0
#> 523  -0.2186765061      0
#> 524  -0.0408533749      0
#> 525  -0.9992666028      0
#> 526   0.1475217993      0
#> 527   0.5591140536      0
#> 528   0.6002019889      0
#> 529  -1.0726146984      0
#> 530  -0.5647159320      0
#> 531   2.4650132737      0
#> 532  -0.8846157496      0
#> 533   0.1837985474      0
#> 534  -0.1060463694      0
#> 535   0.4645933919      0
#> 536   0.0055141156      0
#> 537   1.4725352315      0
#> 538  -0.3605958046      0
#> 539   0.1643795185      0
#> 540   0.0961161214      0
#> 541  -0.4527516010      0
#> 542  -1.6632720526      0
#> 543   0.4328514289      0
#> 544   1.1417698601      0
#> 545   0.5380975080      0
#> 546   0.0141355361      0
#> 547   0.4330982648      0
#> 548  -0.5649218001      0
#> 549  -0.6144507503      0
#> 550  -0.2208972186      0
#> 551   0.1968649965      0
#> 552   1.8243385429      0
#> 553   0.4996744384      0
#> 554   0.5566460683      0
#> 555   0.6922765205      0
#> 556   0.3938625683      0
#> 557   0.9735892240      0
#> 558   0.6165639906      0
#> 559  -1.5434980951      0
#> 560  -0.8987170462      0
#> 561  -0.2166311336      0
#> 562  -1.5070698808      0
#> 563  -0.9365964208      0
#> 564   2.0806856202      0
#> 565   0.2818378753      0
#> 566   0.4970861146      0
#> 567  -2.1200083305      0
#> 568  -0.0703406363      0
#> 569   0.1966989500      0
#> 570  -0.2677151067      0
#> 571  -1.0807926766      0
#> 572  -1.0973446640      0
#> 573   0.1389836975      0
#> 574  -0.9208018387      0
#> 575  -1.5295770232      0
#> 576  -1.4423183264      0
#> 577   0.3518996590      0
#> 578   0.9894768883      0
#> 579  -1.1925041041      0
#> 580   2.3633071379      0
#> 581   1.9875054002      0
#> 582   0.3257709037      0
#> 583   0.8455033543      0
#> 584  -1.1124571045      0
#> 585   1.8640433259      0
#> 586  -0.1607657107      0
#> 587  -0.3530361685      0
#> 588   0.8459714330      0
#> 589   1.1078852027      0
#> 590   2.9275868216      0
#> 591   2.5180132984      0
#> 592  -0.4705193139      0
#> 593  -0.1256621716      0
#> 594  -0.5613266032      0
#> 595   0.8231678015      0
#> 596   0.6222005219      0
#> 597  -0.7730406897      0
#> 598   0.3464437891      0
#> 599   0.7646513250      0
#> 600   0.1814476524      0
#> 601  -1.3336516352      0
#> 602  -0.6454712172      0
#> 603   0.2097679185      0
#> 604  -0.2433621565      0
#> 605  -1.1057539391      0
#> 606  -0.8985566391      0
#> 607   0.1901280779      0
#> 608  -0.1696513511      0
#> 609   1.2300354715      0
#> 610   0.2692057425      0
#> 611   0.7446738356      0
#> 612   0.5776955297      0
#> 613   0.2945233226      0
#> 614   0.3483720565      0
#> 615  -0.3392387819      0
#> 616   2.2386878308      0
#> 617   0.0961553549      0
#> 618   0.7682506708      0
#> 619   0.3353606683      0
#> 620  -0.0825405504      0
#> 621   0.3976927917      0
#> 622  -0.0291976190      0
#> 623  -0.9441511292      0
#> 624   1.2490028164      0
#> 625   0.3066713108      0
#> 626   0.8686836217      0
#> 627   0.3981888044      0
#> 628  -2.0170059980      0
#> 629  -0.5406063913      0
#> 630   0.2047624756      0
#> 631  -1.1198771178      0
#> 632  -0.1230073025      0
#> 633   1.1607541801      0
#> 634   0.2199717476      0
#> 635   0.4562775393      0
#> 636  -0.5791951707      0
#> 637   0.5065188274      0
#> 638  -0.8567953236      0
#> 639  -1.2767535053      0
#> 640  -1.0993045252      0
#> 641  -1.7962038392      0
#> 642   0.8897340496      0
#> 643  -1.0589552386      0
#> 644   2.2084929138      0
#> 645  -0.9858982143      0
#> 646  -0.4577844162      0
#> 647   0.0884393242      0
#> 648   0.9092502155      0
#> 649   0.8257916254      0
#> 650  -0.9860311642      0
#> 651  -0.9963959984      0
#> 652  -0.3204717401      0
#> 653   0.8212155583      0
#> 654   1.0513707455      0
#> 655  -1.1078468814      0
#> 656  -0.8136478250      0
#> 657   0.5510207017      0
#> 658   1.5243601943      0
#> 659  -0.0715061840      0
#> 660  -0.2636309051      0
#> 661   2.2280375891      0
#> 662  -0.0498608088      0
#> 663  -0.5814440853      0
#> 664   0.1082122935      0
#> 665   2.0044670201      0
#> 666  -0.0860193974      0
#> 667   0.0921725006      0
#> 668   0.2668782828      0
#> 669  -0.8169909751      0
#> 670  -0.1552066935      0
#> 671  -0.5250048192      0
#> 672   0.4468450815      0
#> 673   1.9043495385      0
#> 674   1.2248177143      0
#> 675   0.2962838262      0
#> 676  -0.7926247653      0
#> 677  -1.7316058835      0
#> 678   0.0241661430      0
#> 679  -0.8115980919      0
#> 680   0.1809926227      0
#> 681  -2.2085269229      0
#> 682   0.6185072145      0
#> 683   0.0773300388      0
#> 684  -0.4162302571      0
#> 685   0.0530562030      0
#> 686   1.5595120827      0
#> 687  -0.4319266927      0
#> 688   0.1697442573      0
#> 689   0.1290019095      0
#> 690   1.4795107256      0
#> 691  -0.3197019461      0
#> 692   2.1304644523      0
#> 693  -1.1700329688      0
#> 694  -0.1914412900      0
#> 695  -2.3783011922      0
#> 696  -1.0624307313      0
#> 697  -0.8661299729      0
#> 698  -0.5318271362      0
#> 699   1.6667116694      0
#> 700  -2.0041613739      0
#> 701   0.3284621982      0
#> 702  -0.7027085870      0
#> 703   0.4448439346      0
#> 704  -0.2231244603      0
#> 705  -0.2787385811      0
#> 706  -0.6467888708      0
#> 707   0.2608714213      0
#> 708   0.0883272415      0
#> 709   0.8471667047      0
#> 710   0.0809066865      0
#> 711  -1.5760902782      0
#> 712  -0.6073910661      0
#> 713  -0.8200797184      0
#> 714  -2.3908152805      0
#> 715  -0.5407595387      0
#> 716  -0.9932743575      0
#> 717  -1.6321953418      0
#> 718   1.0960007707      0
#> 719   1.6504558876      0
#> 720  -1.2591858219      0
#> 721   0.2267356555      0
#> 722   1.2456204289      0
#> 723  -0.3246034815      0
#> 724   0.8051072369      0
#> 725   0.7025377370      0
#> 726  -0.1488334176      0
#> 727  -0.1889095792      0
#> 728  -0.0184959691      0
#> 729  -1.6750037138      0
#> 730   0.5984271485      0
#> 731  -0.1651176619      0
#> 732   0.4027177818      0
#> 733   0.8873286344      0
#> 734  -0.2481718068      0
#> 735   1.4437398316      0
#> 736   0.0829204791      0
#> 737   1.0257690349      0
#> 738  -1.1824446661      0
#> 739   0.2552012527      0
#> 740  -0.9186743318      0
#> 741  -0.3809354206      0
#> 742   2.9489704453      0
#> 743  -1.4393128474      0
#> 744  -0.7101040055      0
#> 745   0.9715841677      0
#> 746   0.5740010010      0
#> 747  -0.5396257250      0
#> 748   0.2324672927      0
#> 749   0.5448624960      0
#> 750   1.5149158840      0
#> 751  -0.6283642649      0
#> 752  -0.3356217943      0
#> 753   1.1843770002      0
#> 754  -0.1784735150      0
#> 755  -0.0532638560      0
#> 756  -0.0685963534      0
#> 757  -0.4644662317      0
#> 758  -1.5027632907      0
#> 759  -0.5848507344      0
#> 760  -1.1971362517      0
#> 761   0.0523044275      0
#> 762  -0.3212556386      0
#> 763   0.1377255951      0
#> 764   0.4840337824      0
#> 765  -0.1368837349      0
#> 766  -0.4335844711      0
#> 767  -1.8406257380      0
#> 768   1.3609620794      0
#> 769   1.1628676984      0
#> 770  -0.0733241186      0
#> 771   0.2802588902      0
#> 772  -2.4835848359      0
#> 773  -0.8195883943      0
#> 774  -0.5419666166      0
#> 775   0.1762123560      0
#> 776  -0.9876228885      0
#> 777   1.6483057899      0
#> 778   0.9151312703      0
#> 779   0.4880150812      0
#> 780  -0.9374407919      0
#> 781   0.3882993926      0
#> 782   0.3069548356      0
#> 783  -1.2635600249      0
#> 784   0.0224966007      0
#> 785   0.7273676680      0
#> 786  -0.5588384725      0
#> 787  -1.0686523505      0
#> 788   0.3918027894      0
#> 789  -0.6687279053      0
#> 790  -1.0230115457      0
#> 791   0.4200669195      0
#> 792  -0.9771537621      0
#> 793  -0.8133535202      0
#> 794  -1.5858693490      0
#> 795   0.0286849326      0
#> 796   0.3949368316      0
#> 797   0.7838092663      0
#> 798  -0.6591170367      0
#> 799  -0.2490129922      0
#> 800   1.1512326914      0
#> 801   0.1146057072      0
#> 802  -0.4552780962      0
#> 803  -0.4108347796      0
#> 804  -0.2097071806      0
#> 805   0.2968908299      0
#> 806  -0.3494944509      0
#> 807   1.5273053254      0
#> 808  -0.3944215272      0
#> 809  -0.1265355542      0
#> 810   1.4001463052      0
#> 811   0.2138112270      0
#> 812  -0.8089600082      0
#> 813   0.7553746173      0
#> 814   0.2696073293      0
#> 815   0.6942332564      0
#> 816   0.8659918048      0
#> 817   0.6328508883      0
#> 818   0.0604164501      0
#> 819   0.1691610591      0
#> 820  -0.9800951963      0
#> 821   1.6033577317      0
#> 822   1.6075185685      0
#> 823   2.0942073199      0
#> 824   2.0443369692      0
#> 825   0.8871151363      0
#> 826  -1.6993621211      0
#> 827  -0.0099414259      0
#> 828  -0.4975873501      0
#> 829  -0.1831736478      0
#> 830   0.9890107813      0
#> 831   0.6011315831      0
#> 832   0.5168068453      0
#> 833  -0.3575664170      0
#> 834   0.6034728531      0
#> 835   0.2605577777      0
#> 836   1.0545295955      0
#> 837  -0.7801961955      0
#> 838   0.2021667804      0
#> 839  -1.5081502243      0
#> 840   1.1505081504      0
#> 841  -0.0253024450      0
#> 842  -1.2645267535      0
#> 843  -0.5220072962      0
#> 844   0.7297040255      0
#> 845   0.0292367391      0
#> 846   1.1260180941      0
#> 847   0.6461052237      0
#> 848  -0.1724229830      0
#> 849  -1.4231585058      0
#> 850   0.6444475843      0
#> 851  -0.5969702087      0
#> 852   0.1458192532      0
#> 853   0.4413112091      0
#> 854   0.6185563385      0
#> 855   0.8556974452      0
#> 856  -0.0454695493      0
#> 857  -0.6767825084      0
#> 858  -0.6413305968      0
#> 859   0.3066512798      0
#> 860  -0.3888386630      0
#> 861  -1.0891369740      0
#> 862  -0.5120778989      0
#> 863   1.5105128162      0
#> 864  -0.0423166012      0
#> 865  -0.0457991701      0
#> 866   1.1648031177      0
#> 867  -1.4910730058      0
#> 868   1.4300519264      0
#> 869   0.7118828687      0
#> 870   1.0599002050      0
#> 871  -0.2141085800      0
#> 872   1.6914606452      0
#> 873   0.9652167566      0
#> 874  -0.4449281901      0
#> 875  -0.0968577271      0
#> 876   0.8970685338      0
#> 877   0.1322140633      0
#> 878   0.0169258919      0
#> 879  -0.8824387234      0
#> 880   1.5648649436      0
#> 881  -0.6451756126      0
#> 882   0.3626027211      0
#> 883  -1.5319201724      0
#> 884   1.0284527637      0
#> 885  -0.1103738859      0
#> 886   0.2670855274      0
#> 887  -0.1301938009      0
#> 888   0.0148269769      0
#> 889   0.9116064131      0
#> 890  -1.1224053182      0
#> 891  -0.8068740170      0
#> 892   0.6549063433      0
#> 893   0.5188566329      0
#> 894   0.4185187780      0
#> 895   0.2807989873      0
#> 896  -1.4494432459      0
#> 897  -0.7159433653      0
#> 898  -0.4782561984      0
#> 899  -2.0102350976      0
#> 900  -0.0013335363      0
#> 901   0.1461014982      0
#> 902   0.0074583424      0
#> 903   0.1147413854      0
#> 904  -3.0227710010      0
#> 905   1.7026836898      1
#> 906  -0.1004744181      1
#> 907  -0.1272217233      1
#> 908   0.3302549307      1
#> 909   1.6332902267      1
#> 910   0.4353718153      1
#> 911   0.6000766643      1
#> 912   1.1343169162      1
#> 913   0.7643104143      1
#> 914   0.8666023510      1
#> 915   0.3529742090      1
#> 916   0.1014814382      1
#> 917   0.4920087714      1
#> 918  -1.0179981506      1
#> 919  -0.6226491793      1
#> 920   2.1940076821      1
#> 921  -0.4323895794      1
#> 922   1.4941342026      1
#> 923   1.2226807353      1
#> 924   1.2947054396      1
#> 925   1.3268323987      1
#> 926  -0.7308537484      1
#> 927   0.7800648887      1
#> 928   0.4191321942      1
#> 929   1.3260454423      1
#> 930   1.2286538040      1
#> 931  -1.4396229385      1
#> 932   1.1790544376      1
#> 933   1.5801739343      1
#> 934   1.3571404904      1
#> 935   0.9775314339      1
#> 936   1.7010488387      1
#> 937   0.1150589673      1
#> 938   0.7011691571      1
#> 939   1.9682299235      1
#> 940   0.9639486691      1
#> 941   1.9879613189      1
#> 942   0.7627871513      1
#> 943   0.8437322618      1
#> 944   0.1999679577      1
#> 945   1.6241664379      1
#> 946   0.2490293208      1
#> 947  -0.7346964197      1
#> 948   0.8999444251      1
#> 949   1.0178100135      1
#> 950   2.3990623799      1
#> 951   1.5300667340      1
#> 952   0.4947363241      1
#> 953   0.0431092029      1
#> 954   0.2477468114      1
#> 955   1.1540568646      1
#> 956   1.0173104517      1
#> 957   0.3358154779      1
#> 958   1.2889486085      1
#> 959  -0.1335068878      1
#> 960   1.7793913383      1
#> 961  -0.3260010932      1
#> 962   1.8018994839      1
#> 963   1.3702145320      1
#> 964   1.4934173823      1
#> 965   2.1199685373      1
#> 966   2.6143240793      1
#> 967  -1.3679568413      1
#> 968   0.2901343533      1
#> 969   0.1503699180      1
#> 970   1.9951687156      1
#> 971   0.4081246210      1
#> 972   1.4938830183      1
#> 973   1.7322972131      1
#> 974  -0.4063984005      1
#> 975  -0.4767339211      1
#> 976   1.7006894047      1
#> 977   2.1045625675      1
#> 978  -0.9333994909      1
#> 979   0.6962118794      1
#> 980   0.9628976881      1
#> 981   0.6565740215      1
#> 982   0.5836401414      1
#> 983   0.9340412016      1
#> 984  -0.2290215652      1
#> 985   1.0510427450      1
#> 986   1.5210706105      1
#> 987   2.0148007238      1
#> 988  -0.0844664962      1
#> 989  -0.8158899997      1
#> 990   1.6781777395      1
#> 991   0.7180028259      1
#> 992   0.8811593161      1
#> 993   1.5063400910      1
#> 994   0.6002914761      1
#> 995   2.9553013650      1
#> 996   0.2621218202      1
#> 997  -0.0533599614      1
#> 998   0.3082620793      1
#> 999   0.8636002801      1
#> 1000  0.8874543792      1