Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.898345725      0
#> 2    -2.315393638      0
#> 3     0.407125194      0
#> 4     0.740216128      0
#> 5    -1.534679665      0
#> 6     0.350990076      0
#> 7    -0.055695805      0
#> 8     0.175457587      0
#> 9    -1.469718916      0
#> 10    0.897543505      0
#> 11   -0.068922837      0
#> 12    0.362537635      0
#> 13    0.667062530      0
#> 14   -1.411025700      0
#> 15    0.297659530      0
#> 16    2.359741338      0
#> 17   -0.494353314      0
#> 18    0.739118598      0
#> 19    0.416317076      0
#> 20   -1.179742470      0
#> 21   -0.205588918      0
#> 22   -1.057292911      0
#> 23    2.040436405      0
#> 24    0.434089607      0
#> 25   -1.046895873      0
#> 26   -1.091645018      0
#> 27   -0.553666106      0
#> 28   -0.291921900      0
#> 29   -0.008678146      0
#> 30   -0.350884224      0
#> 31    0.124741326      0
#> 32    0.657162850      0
#> 33    1.013082287      0
#> 34   -1.284911577      0
#> 35    0.855443972      0
#> 36   -1.095390163      0
#> 37    0.343279418      0
#> 38    0.084822687      0
#> 39   -0.730825078      0
#> 40   -1.073372802      0
#> 41    0.479428363      0
#> 42   -0.534034418      0
#> 43   -0.018992562      0
#> 44   -0.106420216      0
#> 45    1.040072190      0
#> 46   -0.483975542      0
#> 47   -1.178458263      0
#> 48   -1.641902760      0
#> 49   -0.712311479      0
#> 50    0.305748144      0
#> 51    2.054293703      0
#> 52    1.220813866      0
#> 53    1.142830705      0
#> 54    0.750535045      0
#> 55   -1.055780658      0
#> 56   -1.294440100      0
#> 57   -0.961256644      0
#> 58    0.185252617      0
#> 59    0.344120565      0
#> 60    0.928627354      0
#> 61   -0.046832614      0
#> 62   -0.051467728      0
#> 63    0.836426071      0
#> 64    1.180577302      0
#> 65    1.210216272      0
#> 66    0.185388079      0
#> 67    0.081247946      0
#> 68   -1.066461328      0
#> 69   -0.932825133      0
#> 70   -1.291303681      0
#> 71    0.598217354      0
#> 72   -0.976697911      0
#> 73   -1.589664708      0
#> 74   -0.749699285      0
#> 75    0.304786165      0
#> 76    0.030293248      0
#> 77    0.225884569      0
#> 78   -1.147562298      0
#> 79    0.649268905      0
#> 80   -1.197148445      0
#> 81   -0.770079700      0
#> 82   -1.501074362      0
#> 83   -0.757771082      0
#> 84   -0.512539394      0
#> 85    0.659740728      0
#> 86    0.430153901      0
#> 87    0.047349013      0
#> 88    0.620160947      0
#> 89   -0.504590020      0
#> 90    0.121101179      0
#> 91   -2.060292642      0
#> 92   -0.402266569      0
#> 93   -0.503314948      0
#> 94    1.098918470      0
#> 95   -1.939055005      0
#> 96   -2.167465069      0
#> 97   -0.942577220      0
#> 98   -0.978325274      0
#> 99   -0.605614154      0
#> 100   0.482338397      0
#> 101   0.304821889      0
#> 102   2.177529887      0
#> 103   0.270626626      0
#> 104  -0.897712593      0
#> 105  -0.256855512      0
#> 106   0.159912471      0
#> 107   0.662001356      0
#> 108  -1.692620828      0
#> 109   0.492863968      0
#> 110  -0.561331417      0
#> 111   2.423164000      0
#> 112   0.088285713      0
#> 113  -0.306313509      0
#> 114  -0.795368205      0
#> 115  -0.579292639      0
#> 116   0.844896857      0
#> 117  -1.939157221      0
#> 118   0.698756787      0
#> 119  -0.545728882      0
#> 120   0.069487734      0
#> 121  -0.378300616      0
#> 122  -1.411793471      0
#> 123   0.712294648      0
#> 124  -0.596743587      0
#> 125   0.806068883      0
#> 126  -0.591995326      0
#> 127  -0.281047231      0
#> 128   0.010475827      0
#> 129  -0.318699457      0
#> 130   1.306419742      0
#> 131  -1.206419969      0
#> 132   1.872339403      0
#> 133   0.407666153      0
#> 134   0.522795341      0
#> 135  -0.965728593      0
#> 136  -0.534412674      0
#> 137  -0.469568943      0
#> 138   0.717046937      0
#> 139   0.295321466      0
#> 140   2.164791818      0
#> 141   1.121335896      0
#> 142  -0.304826104      0
#> 143   0.953917851      0
#> 144  -0.320433784      0
#> 145   0.141223316      0
#> 146   1.657248728      0
#> 147  -2.302017859      0
#> 148   1.352070069      0
#> 149  -0.076268548      0
#> 150  -0.424234089      0
#> 151   1.456532630      0
#> 152   0.382311544      0
#> 153   0.731619565      0
#> 154   1.792484432      0
#> 155   0.387821643      0
#> 156   0.007456457      0
#> 157  -0.368781161      0
#> 158  -1.496826986      0
#> 159   1.063785393      0
#> 160   0.409041545      0
#> 161   1.452959033      0
#> 162   0.214946798      0
#> 163   0.358880620      0
#> 164  -0.661017438      0
#> 165   0.680152497      0
#> 166  -2.520536788      0
#> 167   0.584444171      0
#> 168  -0.350478282      0
#> 169  -0.175788065      0
#> 170  -0.672003882      0
#> 171  -0.278998141      0
#> 172   1.334624950      0
#> 173   0.404111269      0
#> 174  -1.759891698      0
#> 175  -0.463314528      0
#> 176  -0.307210676      0
#> 177  -0.005480345      0
#> 178   0.842798480      0
#> 179   0.958428107      0
#> 180  -1.219468171      0
#> 181  -0.253577476      0
#> 182  -1.554793388      0
#> 183  -0.868315560      0
#> 184   0.316674570      0
#> 185  -0.936491901      0
#> 186   0.948955881      0
#> 187  -1.658707240      0
#> 188  -0.283535770      0
#> 189   0.188846386      0
#> 190   0.294274423      0
#> 191   0.691944947      0
#> 192  -0.654673450      0
#> 193  -1.199370405      0
#> 194  -0.321267444      0
#> 195   1.740978771      0
#> 196   1.328958926      0
#> 197   0.997889460      0
#> 198  -1.502635890      0
#> 199  -0.428208949      0
#> 200  -2.218142387      0
#> 201  -0.685281354      0
#> 202   0.009516541      0
#> 203   2.034202759      0
#> 204   0.466124410      0
#> 205  -0.165839104      0
#> 206  -0.475724644      0
#> 207   1.314296977      0
#> 208   1.707548338      0
#> 209  -1.159862195      0
#> 210  -0.776771993      0
#> 211   0.229947975      0
#> 212   0.395329974      0
#> 213   3.303866054      0
#> 214  -0.297718002      0
#> 215  -1.825075946      0
#> 216   0.840066342      0
#> 217   1.767938348      0
#> 218   0.037066796      0
#> 219   0.529728773      0
#> 220   0.560227216      0
#> 221   0.470949139      0
#> 222   0.121441810      0
#> 223   0.070489921      0
#> 224   0.105579735      0
#> 225   0.257819037      0
#> 226   2.001435818      0
#> 227   0.145764540      0
#> 228  -0.519361672      0
#> 229   0.011084895      0
#> 230  -0.635302261      0
#> 231   0.288727042      0
#> 232   0.966650040      0
#> 233  -0.651322108      0
#> 234  -0.266889501      0
#> 235   0.693187407      0
#> 236  -0.654858245      0
#> 237   0.319042321      0
#> 238   0.262879951      0
#> 239   0.049356723      0
#> 240   0.270636401      0
#> 241  -0.595123261      0
#> 242  -0.227450675      0
#> 243  -0.036336459      0
#> 244  -1.086590018      0
#> 245  -2.812610731      0
#> 246  -1.113370843      0
#> 247   0.253128481      0
#> 248  -0.138026677      0
#> 249   0.023125533      0
#> 250  -0.145554797      0
#> 251  -0.854859459      0
#> 252  -0.507076957      0
#> 253  -2.387937675      0
#> 254   0.901976533      0
#> 255   1.180883326      0
#> 256   0.010800752      0
#> 257   0.137392383      0
#> 258   0.241649420      0
#> 259  -0.449251300      0
#> 260  -1.961986729      0
#> 261   0.327016094      0
#> 262   1.551098668      0
#> 263  -0.729842299      0
#> 264  -0.509379888      0
#> 265   0.513922612      0
#> 266   0.659940769      0
#> 267   1.148473211      0
#> 268  -0.124932433      0
#> 269  -1.351009296      0
#> 270  -0.658596584      0
#> 271  -1.211133174      0
#> 272   0.249364012      0
#> 273  -2.463104967      0
#> 274   0.863183011      0
#> 275   0.463576702      0
#> 276   0.371376392      0
#> 277   0.788766385      0
#> 278   0.771057202      0
#> 279   1.434000155      0
#> 280  -0.935427404      0
#> 281   0.555634514      0
#> 282  -0.986115138      0
#> 283   0.278758500      0
#> 284  -0.181478763      0
#> 285   0.835800261      0
#> 286  -1.090514790      0
#> 287   1.392178297      0
#> 288   0.236585621      0
#> 289   1.077578244      0
#> 290  -0.354108030      0
#> 291  -1.559199801      0
#> 292   1.348578498      0
#> 293  -0.274609542      0
#> 294   0.651182545      0
#> 295  -0.651455169      0
#> 296  -0.514785830      0
#> 297   1.172508781      0
#> 298  -0.279680506      0
#> 299   1.824349320      0
#> 300  -0.609539828      0
#> 301  -1.138787998      0
#> 302  -0.894551086      0
#> 303  -0.358040016      0
#> 304   0.690011365      0
#> 305   0.650645073      0
#> 306   0.316952055      0
#> 307   1.582074123      0
#> 308  -0.894231307      0
#> 309  -0.436662886      0
#> 310  -0.622260592      0
#> 311   0.168925672      0
#> 312  -0.312511753      0
#> 313   0.948146674      0
#> 314  -1.377346890      0
#> 315  -0.362916603      0
#> 316  -0.759391517      0
#> 317  -1.134912167      0
#> 318  -1.016700056      0
#> 319  -0.924800352      0
#> 320   0.876576675      0
#> 321   0.333667747      0
#> 322  -0.975168372      0
#> 323   1.064900571      0
#> 324   0.862957629      0
#> 325  -0.558572144      0
#> 326  -0.180574764      0
#> 327  -0.420161518      0
#> 328  -0.225332246      0
#> 329  -0.354094521      0
#> 330   0.001151694      0
#> 331  -0.343529662      0
#> 332  -1.436082574      0
#> 333  -0.654787319      0
#> 334  -0.951712625      0
#> 335  -0.184443273      0
#> 336  -0.343815214      0
#> 337  -0.017092310      0
#> 338   0.966409239      0
#> 339  -0.359887231      0
#> 340   0.311292895      0
#> 341  -1.576821284      0
#> 342  -0.198046921      0
#> 343  -0.438424619      0
#> 344   1.397638678      0
#> 345   1.311152051      0
#> 346   0.564913153      0
#> 347   1.059290328      0
#> 348   1.211216723      0
#> 349   0.763112133      0
#> 350   0.180677991      0
#> 351  -0.361834230      0
#> 352  -0.677551129      0
#> 353   0.451028621      0
#> 354  -0.283177469      0
#> 355   0.661198662      0
#> 356   0.891371338      0
#> 357   0.134109221      0
#> 358   1.891131087      0
#> 359  -0.616090173      0
#> 360  -1.531799107      0
#> 361   0.461544239      0
#> 362   0.424170461      0
#> 363  -0.897618913      0
#> 364   0.829430622      0
#> 365  -0.731078100      0
#> 366  -1.460343196      0
#> 367  -0.941406697      0
#> 368   0.085319352      0
#> 369   0.491763814      0
#> 370   0.142338223      0
#> 371   0.268535625      0
#> 372   0.944594955      0
#> 373   0.810318476      0
#> 374   0.217805218      0
#> 375  -0.108983389      0
#> 376   0.704436084      0
#> 377  -1.011826512      0
#> 378  -0.131741542      0
#> 379  -0.887442512      0
#> 380  -0.222619162      0
#> 381  -0.222598066      0
#> 382  -0.984017295      0
#> 383   1.044611823      0
#> 384  -0.372125081      0
#> 385  -1.740404385      0
#> 386  -1.009429813      0
#> 387   0.837187781      0
#> 388   0.074228988      0
#> 389  -1.186025828      0
#> 390   0.339639776      0
#> 391   0.795822767      0
#> 392   0.449578903      0
#> 393   0.870330789      0
#> 394  -0.901523973      0
#> 395  -0.067275405      0
#> 396  -0.918950515      0
#> 397   0.357398749      0
#> 398  -0.671111533      0
#> 399  -0.833052973      0
#> 400  -1.088011325      0
#> 401   1.830694039      0
#> 402  -0.179659468      0
#> 403   0.847116051      0
#> 404  -0.918084775      0
#> 405  -1.040034173      0
#> 406  -0.645452401      0
#> 407  -0.777064688      0
#> 408  -1.027744580      0
#> 409  -1.008533540      0
#> 410   0.680026721      0
#> 411  -1.038671436      0
#> 412   0.457875007      0
#> 413  -0.588664034      0
#> 414   0.772712553      0
#> 415  -0.269999925      0
#> 416   0.547631066      0
#> 417  -0.892762444      0
#> 418   1.720707865      0
#> 419   0.444208333      0
#> 420  -0.510344643      0
#> 421   1.491152749      0
#> 422   0.309688163      0
#> 423  -1.083072056      0
#> 424   0.459368729      0
#> 425   0.116700043      0
#> 426  -0.539327347      0
#> 427  -0.309600141      0
#> 428   0.385676956      0
#> 429   1.080139390      0
#> 430  -0.859442926      0
#> 431  -0.355238044      0
#> 432  -1.529408400      0
#> 433   0.292375905      0
#> 434  -1.507132245      0
#> 435  -1.470805832      0
#> 436   0.768537706      0
#> 437  -0.991085280      0
#> 438  -0.792874362      0
#> 439  -1.207153127      0
#> 440   0.707742829      0
#> 441  -0.103211261      0
#> 442   0.516841249      0
#> 443  -1.266031002      0
#> 444  -0.480473972      0
#> 445  -0.959425422      0
#> 446  -0.153565955      0
#> 447   2.484774951      0
#> 448  -0.825430056      0
#> 449  -1.405334692      0
#> 450  -0.143817842      0
#> 451   0.197907506      0
#> 452   0.490270338      0
#> 453   0.736295601      0
#> 454  -0.069544712      0
#> 455   0.447776129      0
#> 456   0.422019133      0
#> 457  -0.811486144      0
#> 458  -1.120673512      0
#> 459   0.637193983      0
#> 460   0.848528680      0
#> 461  -0.411927741      0
#> 462  -1.140175963      0
#> 463   0.181454421      0
#> 464  -1.085616828      0
#> 465  -1.547546929      0
#> 466  -0.416851822      0
#> 467  -0.138225403      0
#> 468  -0.095670458      0
#> 469   0.813036064      0
#> 470  -0.185567980      0
#> 471   1.016101618      0
#> 472   2.355456674      0
#> 473  -1.064822191      0
#> 474   0.728256259      0
#> 475  -0.739152922      0
#> 476  -0.260318122      0
#> 477  -1.719118899      0
#> 478   0.911967866      0
#> 479   0.510492660      0
#> 480  -1.090886634      0
#> 481   0.369128253      0
#> 482  -0.794989631      0
#> 483  -0.129495844      0
#> 484  -0.523310997      0
#> 485   0.941525444      0
#> 486   0.580077091      0
#> 487  -0.694117352      0
#> 488   1.032767784      0
#> 489   0.511109266      0
#> 490   0.707891440      0
#> 491   1.541414611      0
#> 492   0.155499160      0
#> 493  -0.364711594      0
#> 494  -0.636260606      0
#> 495   1.068407125      0
#> 496   1.168029888      0
#> 497  -1.166931798      0
#> 498  -0.759357874      0
#> 499   0.004081023      0
#> 500  -0.305952791      0
#> 501   1.162723879      0
#> 502   0.563992876      0
#> 503   0.333912577      0
#> 504   0.772827282      0
#> 505  -0.226264606      0
#> 506   0.754497659      0
#> 507  -0.057999022      0
#> 508  -1.210109315      0
#> 509   0.411943924      0
#> 510   2.278164294      0
#> 511   0.902810342      0
#> 512  -0.955826326      0
#> 513   0.478161750      0
#> 514  -1.613552701      0
#> 515  -0.814444164      0
#> 516   0.889137554      0
#> 517   1.072790187      0
#> 518   0.004966254      0
#> 519   0.519210772      0
#> 520  -2.451437335      0
#> 521   0.267611863      0
#> 522   0.291921414      0
#> 523  -0.315845688      0
#> 524   1.295486395      0
#> 525  -0.831264404      0
#> 526   2.636394289      0
#> 527   0.234620582      0
#> 528   0.390497666      0
#> 529   1.069330416      0
#> 530  -0.288075035      0
#> 531  -1.047338047      0
#> 532   0.110541867      0
#> 533  -0.122050928      0
#> 534  -0.604514273      0
#> 535  -0.099619902      0
#> 536  -0.017566442      0
#> 537   0.799018335      0
#> 538   0.814801276      0
#> 539   0.661166403      0
#> 540  -0.107715688      0
#> 541   0.827759269      0
#> 542  -0.011251446      0
#> 543   0.522031900      0
#> 544   0.228955340      0
#> 545   0.244436835      0
#> 546   0.440902440      0
#> 547  -1.408085844      0
#> 548   2.632218238      0
#> 549   0.717249301      0
#> 550  -1.239962000      0
#> 551   1.900904764      0
#> 552  -0.529521683      0
#> 553  -0.821841746      0
#> 554   1.181121433      0
#> 555  -0.881782928      0
#> 556   0.337786854      0
#> 557  -0.309526341      0
#> 558  -0.834643641      0
#> 559  -0.213961922      0
#> 560   0.087584494      0
#> 561   0.124110217      0
#> 562   0.084651765      0
#> 563  -1.535214274      0
#> 564  -0.534698186      0
#> 565   1.677171904      0
#> 566  -0.536039452      0
#> 567   0.968052198      0
#> 568   0.152264227      0
#> 569  -0.006299872      0
#> 570   0.715823328      0
#> 571   0.694146349      0
#> 572  -1.453009334      0
#> 573  -1.774744025      0
#> 574  -1.214534916      0
#> 575  -0.227711552      0
#> 576   0.277929398      0
#> 577   0.446466389      0
#> 578  -0.116919915      0
#> 579  -0.330414787      0
#> 580   0.165406386      0
#> 581   0.138193372      0
#> 582   1.001777983      0
#> 583  -0.575441834      0
#> 584  -0.775081714      0
#> 585  -1.913056189      0
#> 586   0.585521044      0
#> 587   1.226965629      0
#> 588  -0.232108733      0
#> 589   1.227231725      0
#> 590  -0.348762534      0
#> 591   2.094886631      0
#> 592   1.779363019      0
#> 593   0.967731395      0
#> 594   0.065806578      0
#> 595   0.117146063      0
#> 596  -0.278608978      0
#> 597   0.346548077      0
#> 598   1.551842764      0
#> 599  -1.715619104      0
#> 600  -0.275827041      0
#> 601  -0.993079556      0
#> 602  -0.163398357      0
#> 603   1.381032840      0
#> 604  -1.137456759      0
#> 605   0.119120877      0
#> 606   0.219322113      0
#> 607  -0.524905602      0
#> 608  -1.008549635      0
#> 609   0.237357322      0
#> 610  -1.453419137      0
#> 611   2.875171343      0
#> 612   0.283417239      0
#> 613  -0.437326011      0
#> 614   0.443808636      0
#> 615   0.646140889      0
#> 616  -1.124908075      0
#> 617  -0.250448585      0
#> 618  -0.855706860      0
#> 619  -1.099643165      0
#> 620  -0.563270768      0
#> 621   0.766358097      0
#> 622   1.477007763      0
#> 623  -2.272008507      0
#> 624   2.536031085      0
#> 625  -0.297697954      0
#> 626   1.359119238      0
#> 627  -0.197480276      0
#> 628  -2.951325534      0
#> 629   0.895892954      0
#> 630  -0.577362867      0
#> 631   0.813583773      0
#> 632   1.966256912      0
#> 633  -2.646516785      0
#> 634  -0.040830507      0
#> 635   1.132637321      0
#> 636  -0.575692713      0
#> 637  -0.253798460      0
#> 638  -1.993795062      0
#> 639   3.778810452      0
#> 640  -0.630050815      0
#> 641   0.803747367      0
#> 642   1.726454575      0
#> 643   0.296554668      0
#> 644  -0.973042234      0
#> 645  -0.610786351      0
#> 646   1.081674395      0
#> 647  -0.165247216      0
#> 648  -0.792128832      0
#> 649   1.809916650      0
#> 650  -0.127262637      0
#> 651  -0.807093196      0
#> 652   1.083782073      0
#> 653   0.132459408      0
#> 654   0.264465602      0
#> 655   0.030050443      0
#> 656   1.592126779      0
#> 657  -2.012322325      0
#> 658  -1.539214703      0
#> 659  -0.616293553      0
#> 660   1.640489758      0
#> 661   1.980499604      0
#> 662   0.049482417      0
#> 663  -0.397930138      0
#> 664  -0.595022651      0
#> 665  -1.252317595      0
#> 666   0.774466469      0
#> 667  -0.927607460      0
#> 668  -1.512369552      0
#> 669  -0.301488842      0
#> 670   0.119162009      0
#> 671  -2.126056947      0
#> 672  -0.170012037      0
#> 673   0.197546313      0
#> 674   1.456853231      0
#> 675   0.463486941      0
#> 676   1.252598382      0
#> 677   0.810625454      0
#> 678   0.039636787      0
#> 679   1.144440656      0
#> 680  -0.949945286      0
#> 681   0.600212006      0
#> 682  -0.204281227      0
#> 683  -1.103513358      0
#> 684  -1.480983752      0
#> 685   2.440720364      0
#> 686   1.481852448      0
#> 687   0.724901335      0
#> 688   0.572290119      0
#> 689  -0.701221007      0
#> 690  -0.996113300      0
#> 691   0.218046634      0
#> 692  -1.573265623      0
#> 693  -1.043031949      0
#> 694   0.181760305      0
#> 695   2.603650468      0
#> 696   0.446315044      0
#> 697  -1.444276008      0
#> 698   0.343643034      0
#> 699  -0.364896850      0
#> 700   1.360165455      0
#> 701  -0.857136509      0
#> 702   0.678554948      0
#> 703   0.340538834      0
#> 704  -1.600175158      0
#> 705  -1.457242343      0
#> 706   0.699537575      0
#> 707   0.822054628      0
#> 708   1.261981224      0
#> 709  -0.255921190      0
#> 710  -0.086448917      0
#> 711   0.493470992      0
#> 712  -0.621596250      0
#> 713  -1.044389960      0
#> 714  -0.965908484      0
#> 715  -2.628560417      0
#> 716  -0.209717792      0
#> 717   0.477679715      0
#> 718   0.530256711      0
#> 719   1.211876676      0
#> 720  -0.873957052      0
#> 721  -0.209036928      0
#> 722  -1.362524269      0
#> 723  -0.113256052      0
#> 724   0.665974023      0
#> 725  -0.384287917      0
#> 726  -0.430088051      0
#> 727  -0.611213770      0
#> 728   0.416908789      0
#> 729  -0.740888669      0
#> 730   3.214203025      0
#> 731  -0.803406367      0
#> 732   1.206622851      0
#> 733  -1.590609643      0
#> 734  -1.939900399      0
#> 735   2.315326679      0
#> 736  -0.937689446      0
#> 737  -0.946755595      0
#> 738  -0.316225774      0
#> 739   1.428863113      0
#> 740  -1.037747207      0
#> 741   0.375250049      0
#> 742   0.256864232      0
#> 743   0.271022985      0
#> 744  -1.033891036      0
#> 745   1.505578602      0
#> 746   0.432075411      0
#> 747   1.166880527      0
#> 748   0.362238140      0
#> 749  -1.159037940      0
#> 750   0.116658100      0
#> 751  -0.286397370      0
#> 752   0.910679799      0
#> 753   1.017906664      0
#> 754  -0.225748264      0
#> 755  -0.544435485      0
#> 756  -1.437092348      0
#> 757   0.545216899      0
#> 758  -0.915774894      0
#> 759  -0.657971932      0
#> 760  -0.626602267      0
#> 761  -1.062629286      0
#> 762  -0.973853900      0
#> 763  -2.554371338      0
#> 764  -1.888604062      0
#> 765   0.102049112      0
#> 766   0.936635276      0
#> 767   0.139728797      0
#> 768  -0.151185265      0
#> 769   0.427983258      0
#> 770  -0.141587043      0
#> 771  -0.187930171      0
#> 772   0.146623369      0
#> 773   0.947085417      0
#> 774   0.109988326      0
#> 775  -1.085927421      0
#> 776  -2.105307290      0
#> 777  -1.038621297      0
#> 778   1.126745333      0
#> 779  -1.764312865      0
#> 780   1.013524254      0
#> 781  -0.247407863      0
#> 782   1.481962080      0
#> 783  -1.567993335      0
#> 784  -0.815854901      0
#> 785  -0.053832863      0
#> 786   0.112187075      0
#> 787   0.120833338      0
#> 788   1.405157659      0
#> 789   0.460284832      0
#> 790  -0.161893586      0
#> 791  -0.293369765      0
#> 792  -0.132904605      0
#> 793  -1.321043554      0
#> 794  -1.315704925      0
#> 795  -0.538971416      0
#> 796   0.523210035      0
#> 797  -0.701838265      0
#> 798  -0.831193742      0
#> 799   0.251789292      0
#> 800  -0.379408952      0
#> 801  -3.094605792      0
#> 802  -0.035977006      0
#> 803   0.727130479      0
#> 804   2.128717408      0
#> 805  -0.538828424      0
#> 806   0.368708493      0
#> 807   0.562654923      0
#> 808   1.348922552      0
#> 809   0.442852136      0
#> 810  -0.047867444      0
#> 811   0.119387631      0
#> 812   2.415548858      0
#> 813  -1.060696301      0
#> 814  -0.366315362      0
#> 815   0.039164475      0
#> 816   0.127653461      0
#> 817  -0.448714259      0
#> 818  -0.178097784      0
#> 819   0.337879644      0
#> 820   0.031780998      0
#> 821   0.692298731      0
#> 822  -0.279450433      0
#> 823   1.140703786      0
#> 824   1.649126124      0
#> 825   0.564947101      0
#> 826  -0.318646013      0
#> 827   0.159659778      0
#> 828  -0.081810520      0
#> 829  -0.738484811      0
#> 830  -0.695915581      0
#> 831  -1.224112149      0
#> 832   3.377175304      0
#> 833   0.599428804      0
#> 834  -2.131724070      0
#> 835   0.848258831      0
#> 836  -0.042910047      0
#> 837  -0.720467011      0
#> 838   0.287413132      0
#> 839  -0.045970787      0
#> 840   1.911942031      0
#> 841   0.544112688      0
#> 842  -1.196671983      0
#> 843  -0.581966409      0
#> 844  -0.239053823      0
#> 845   1.806896245      0
#> 846  -0.473198651      0
#> 847  -0.828833844      0
#> 848  -0.218112421      0
#> 849   0.455137105      0
#> 850  -1.998360298      0
#> 851   0.594770914      0
#> 852  -2.292824155      0
#> 853  -0.222501235      0
#> 854  -1.518190817      0
#> 855   0.348427893      0
#> 856   0.233255610      0
#> 857   0.130937780      0
#> 858   0.106902723      0
#> 859  -0.348896717      0
#> 860   1.091346902      0
#> 861   0.933557606      0
#> 862  -0.486284603      0
#> 863  -0.936499536      0
#> 864  -0.429088843      0
#> 865   0.294890321      0
#> 866   0.289491345      0
#> 867  -0.944819074      0
#> 868   0.412044416      0
#> 869   0.939574827      0
#> 870  -0.043054203      0
#> 871  -1.429990301      0
#> 872   0.615547299      0
#> 873   0.490987427      0
#> 874  -0.234682513      0
#> 875   0.174184304      0
#> 876  -1.274041124      0
#> 877  -0.064210215      0
#> 878  -0.663607163      0
#> 879   0.587792444      0
#> 880   0.225896264      0
#> 881  -0.226116084      0
#> 882  -0.083897905      0
#> 883  -0.271779938      0
#> 884  -0.580599138      0
#> 885   1.158765524      0
#> 886  -0.440625252      0
#> 887  -0.856750790      0
#> 888  -0.153326650      0
#> 889  -0.311004444      0
#> 890  -1.557399993      0
#> 891   0.340272604      0
#> 892   0.024690226      0
#> 893   0.052205992      0
#> 894   0.726957022      0
#> 895  -1.531811581      0
#> 896  -0.583102629      0
#> 897  -1.527283096      0
#> 898   2.088179591      0
#> 899   0.788652586      0
#> 900  -0.361968682      0
#> 901  -0.331238787      0
#> 902   0.901249477      0
#> 903  -1.177099102      0
#> 904  -0.011039532      0
#> 905  -0.359807192      0
#> 906   0.209921917      0
#> 907   1.728926536      0
#> 908  -0.501999702      0
#> 909  -0.039400917      0
#> 910  -0.110670698      0
#> 911  -1.743897312      0
#> 912   2.079980008      0
#> 913   0.315109144      0
#> 914   1.032391854      0
#> 915   0.560106724      0
#> 916   1.288425244      0
#> 917  -0.340442314      0
#> 918   0.940778068      0
#> 919   0.213648497      1
#> 920   1.223435393      1
#> 921   1.658254734      1
#> 922  -1.896017671      1
#> 923   0.337555903      1
#> 924   0.785042262      1
#> 925   0.761383698      1
#> 926   0.697210852      1
#> 927   0.849001963      1
#> 928   0.217210761      1
#> 929   1.539191262      1
#> 930  -0.826295551      1
#> 931  -0.229010729      1
#> 932  -0.112409171      1
#> 933   0.062933915      1
#> 934  -0.257903316      1
#> 935   2.704969877      1
#> 936   1.676416291      1
#> 937   1.279322294      1
#> 938   1.672795344      1
#> 939   0.093434152      1
#> 940   0.480618553      1
#> 941  -0.685300691      1
#> 942  -0.772391441      1
#> 943   0.431957050      1
#> 944   0.107228244      1
#> 945   0.011377007      1
#> 946   1.139851098      1
#> 947   1.335962537      1
#> 948   0.961458501      1
#> 949  -0.209384693      1
#> 950   1.381234097      1
#> 951   1.829383387      1
#> 952   2.585042907      1
#> 953   1.686152389      1
#> 954   1.075793180      1
#> 955   1.183988984      1
#> 956   0.410529154      1
#> 957   0.998904618      1
#> 958   0.282827097      1
#> 959   1.603076482      1
#> 960   0.750951072      1
#> 961  -0.023410749      1
#> 962   2.432583682      1
#> 963   0.553613437      1
#> 964   0.622816671      1
#> 965   0.337474640      1
#> 966   1.119042571      1
#> 967   1.059835200      1
#> 968   4.514286979      1
#> 969   0.604755974      1
#> 970   0.508269792      1
#> 971   0.515587859      1
#> 972  -0.107947668      1
#> 973   1.250424200      1
#> 974   0.052701713      1
#> 975  -0.414431578      1
#> 976  -0.146425182      1
#> 977  -0.014229430      1
#> 978   0.226253004      1
#> 979   1.536076308      1
#> 980   0.114157979      1
#> 981   0.562258869      1
#> 982  -0.673850444      1
#> 983   1.078327162      1
#> 984  -0.403148560      1
#> 985   0.229989656      1
#> 986  -0.143651727      1
#> 987   0.589813804      1
#> 988   0.887944163      1
#> 989   0.255799185      1
#> 990  -0.810905536      1
#> 991   1.093977074      1
#> 992   0.385721174      1
#> 993   1.996952561      1
#> 994  -0.140223807      1
#> 995   1.476226381      1
#> 996   0.807073849      1
#> 997   0.745656744      1
#> 998   1.764839318      1
#> 999   0.656298442      1
#> 1000  0.542802458      1