Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.171359482      0
#> 2    -1.749375828      0
#> 3     1.586283488      0
#> 4    -0.908639251      0
#> 5    -0.154883186      0
#> 6     0.247946159      0
#> 7    -0.123132460      0
#> 8     0.709540238      0
#> 9    -0.752681886      0
#> 10   -1.272604194      0
#> 11    0.055871239      0
#> 12   -0.929073047      0
#> 13   -0.875127992      0
#> 14    0.821129201      0
#> 15   -2.233840865      0
#> 16   -2.822822378      0
#> 17    1.316099789      0
#> 18    0.820885271      0
#> 19   -0.750525155      0
#> 20   -1.758843564      0
#> 21    0.710256760      0
#> 22    0.368804986      0
#> 23    0.049717482      0
#> 24   -0.201255082      0
#> 25   -0.928843129      0
#> 26    1.138298517      0
#> 27    0.690327582      0
#> 28    0.298271153      0
#> 29   -1.083777301      0
#> 30   -0.370125627      0
#> 31    0.496755993      0
#> 32   -0.685197019      0
#> 33   -0.575061197      0
#> 34    0.524449994      0
#> 35   -0.861671615      0
#> 36    0.927934847      0
#> 37    1.040532978      0
#> 38   -0.181196339      0
#> 39   -0.193397410      0
#> 40   -0.125055714      0
#> 41    0.787382323      0
#> 42   -0.489834369      0
#> 43    0.210872792      0
#> 44   -0.342624456      0
#> 45    0.261694938      0
#> 46   -0.469344463      0
#> 47    0.296642984      0
#> 48    0.280251888      0
#> 49   -1.137581059      0
#> 50   -0.436269938      0
#> 51   -1.652188504      0
#> 52    1.205930026      0
#> 53   -0.570569694      0
#> 54    0.695288707      0
#> 55   -0.329053444      0
#> 56   -0.175462815      0
#> 57    0.288981062      0
#> 58    0.401372535      0
#> 59    2.096393428      0
#> 60   -0.720308189      0
#> 61   -0.132879340      0
#> 62    1.615111043      0
#> 63   -0.757552504      0
#> 64    0.028763565      0
#> 65   -0.374013116      0
#> 66    2.268858920      0
#> 67   -0.338055159      0
#> 68    0.761670493      0
#> 69   -0.315976705      0
#> 70    0.945966495      0
#> 71    2.278980540      0
#> 72    0.289611110      0
#> 73   -0.719654862      0
#> 74    0.058427311      0
#> 75    1.131144494      0
#> 76    0.040175828      0
#> 77   -0.145662168      0
#> 78    1.183433717      0
#> 79   -0.516193366      0
#> 80   -0.222361741      0
#> 81   -0.588598100      0
#> 82    0.363686314      0
#> 83   -0.872593622      0
#> 84    0.469664516      0
#> 85   -0.306829712      0
#> 86   -1.071712271      0
#> 87    0.992954331      0
#> 88   -1.075975011      0
#> 89   -0.016260355      0
#> 90    1.573880821      0
#> 91    0.774706121      0
#> 92   -0.077811897      0
#> 93    1.110827836      0
#> 94   -1.535037350      0
#> 95   -2.045667465      0
#> 96   -1.075202463      0
#> 97    0.408997060      0
#> 98   -0.992531995      0
#> 99   -0.417735970      0
#> 100   1.900201263      0
#> 101  -1.193671615      0
#> 102   0.819439821      0
#> 103  -0.819090413      0
#> 104   0.998861199      0
#> 105  -0.132229466      0
#> 106   0.016075479      0
#> 107   0.151307887      0
#> 108  -0.774625205      0
#> 109   0.044597571      0
#> 110  -0.336987532      0
#> 111   0.949434775      0
#> 112   1.436515255      0
#> 113  -1.031632503      0
#> 114   0.781912949      0
#> 115  -0.213167155      0
#> 116  -1.100924747      0
#> 117   0.348087801      0
#> 118  -0.445431400      0
#> 119   0.235193543      0
#> 120  -0.029357303      0
#> 121   0.763637552      0
#> 122   0.005447836      0
#> 123  -0.579431217      0
#> 124   1.575091770      0
#> 125  -0.567045825      0
#> 126   0.361985651      0
#> 127  -0.099757197      0
#> 128   2.398799810      0
#> 129   1.265572596      0
#> 130   1.767995102      0
#> 131  -0.584317320      0
#> 132   0.536593146      0
#> 133  -0.246156471      0
#> 134  -0.561185115      0
#> 135   0.991149785      0
#> 136  -0.982707579      0
#> 137  -0.920867313      0
#> 138   0.331840166      0
#> 139  -0.359164774      0
#> 140   1.181331065      0
#> 141   0.657768083      0
#> 142   0.660742802      0
#> 143   1.068433574      0
#> 144   0.224626096      0
#> 145   1.241389231      0
#> 146  -0.117053631      0
#> 147  -0.858330264      0
#> 148  -0.923367419      0
#> 149   0.278299221      0
#> 150   0.491313921      0
#> 151   0.414740637      0
#> 152  -1.014182214      0
#> 153   0.336615125      0
#> 154  -0.491380896      0
#> 155   0.563652884      0
#> 156  -0.778836489      0
#> 157  -0.925118992      0
#> 158   1.072220850      0
#> 159   0.189375503      0
#> 160   0.526933117      0
#> 161   1.677996631      0
#> 162   0.842445633      0
#> 163   0.204204347      0
#> 164  -1.106041860      0
#> 165   1.455424765      0
#> 166  -0.439622948      0
#> 167   0.181227138      0
#> 168  -0.479048333      0
#> 169   1.882833052      0
#> 170   0.809502020      0
#> 171  -0.849645971      0
#> 172  -0.472273729      0
#> 173  -0.954536022      0
#> 174   0.215934803      0
#> 175   0.733608374      0
#> 176   0.721731408      0
#> 177  -0.121200681      0
#> 178   0.846101304      0
#> 179   0.631860627      0
#> 180   0.165669235      0
#> 181   0.731528473      0
#> 182  -1.095174559      0
#> 183   0.753917545      0
#> 184   2.347108389      0
#> 185   1.227470240      0
#> 186  -3.182966741      0
#> 187  -0.914537054      0
#> 188   0.309226608      0
#> 189   0.656887515      0
#> 190  -0.401325338      0
#> 191   0.251932813      0
#> 192  -1.561213964      0
#> 193   1.339664354      0
#> 194  -0.114493725      0
#> 195   2.586195424      0
#> 196  -0.633557795      0
#> 197  -1.267015774      0
#> 198   2.060622038      0
#> 199  -1.718880427      0
#> 200   0.186003877      0
#> 201  -0.420759084      0
#> 202   1.836204127      0
#> 203   2.616346629      0
#> 204   0.214034336      0
#> 205   0.821395849      0
#> 206   0.094252962      0
#> 207  -0.456528564      0
#> 208   0.020509511      0
#> 209   0.254389288      0
#> 210  -2.259797147      0
#> 211   0.417492181      0
#> 212  -0.586731239      0
#> 213  -0.176464383      0
#> 214   0.368065937      0
#> 215   0.733365367      0
#> 216  -0.328828313      0
#> 217   1.235427471      0
#> 218  -0.354852910      0
#> 219   0.407293200      0
#> 220   0.613495440      0
#> 221  -1.478492581      0
#> 222  -0.903886873      0
#> 223  -0.286935243      0
#> 224  -0.479490825      0
#> 225   0.348797381      0
#> 226  -2.341223598      0
#> 227  -1.395228776      0
#> 228  -0.924747272      0
#> 229   1.153100020      0
#> 230  -0.944113254      0
#> 231  -0.228075950      0
#> 232   0.519128956      0
#> 233  -1.113703838      0
#> 234  -0.281311439      0
#> 235   0.285089660      0
#> 236  -0.727021457      0
#> 237   1.227702030      0
#> 238  -0.377490237      0
#> 239   0.133150184      0
#> 240   0.338347659      0
#> 241   0.318498437      0
#> 242  -0.943154462      0
#> 243   0.658134676      0
#> 244   0.120102763      0
#> 245  -0.436903476      0
#> 246   0.453837707      0
#> 247   1.023132287      0
#> 248   0.822757533      0
#> 249  -1.218067947      0
#> 250  -0.491161140      0
#> 251   0.099413471      0
#> 252  -0.446688986      0
#> 253  -0.519252808      0
#> 254   0.962967677      0
#> 255   0.545783887      0
#> 256   1.177177784      0
#> 257  -0.983240476      0
#> 258   0.095548730      0
#> 259  -1.428133303      0
#> 260   0.329412440      0
#> 261  -0.105009030      0
#> 262   0.243565865      0
#> 263   1.008462308      0
#> 264   1.629798761      0
#> 265  -1.291515988      0
#> 266  -1.971313763      0
#> 267  -0.498846735      0
#> 268  -0.467589824      0
#> 269  -0.338961241      0
#> 270   2.350891878      0
#> 271  -1.340930507      0
#> 272   0.515639876      0
#> 273  -0.943436740      0
#> 274   0.666106807      0
#> 275  -0.578343349      0
#> 276   0.287108253      0
#> 277   0.761778326      0
#> 278   0.598235148      0
#> 279  -2.064202232      0
#> 280  -1.258543530      0
#> 281   0.741776962      0
#> 282   1.516982982      0
#> 283   0.929049489      0
#> 284   0.724724237      0
#> 285   0.567409720      0
#> 286  -0.074890030      0
#> 287   1.252165411      0
#> 288   1.217418937      0
#> 289  -0.066036363      0
#> 290   1.667383948      0
#> 291  -1.234709064      0
#> 292  -1.135521113      0
#> 293   0.349484670      0
#> 294  -0.340105962      0
#> 295   0.093284466      0
#> 296  -0.552499176      0
#> 297   0.185823209      0
#> 298   0.899080191      0
#> 299  -0.233574889      0
#> 300  -0.825413952      0
#> 301   1.380809123      0
#> 302   0.615177992      0
#> 303  -0.967006092      0
#> 304  -1.441215476      0
#> 305   1.600266645      0
#> 306  -0.474057117      0
#> 307  -0.418805613      0
#> 308   0.465576675      0
#> 309  -0.377853058      0
#> 310  -0.152781084      0
#> 311  -0.552256412      0
#> 312   0.347290174      0
#> 313   0.740865000      0
#> 314   0.245060095      0
#> 315  -1.248846259      0
#> 316   1.052613340      0
#> 317   1.606099825      0
#> 318  -0.272497776      0
#> 319   1.267693602      0
#> 320  -2.195743606      0
#> 321  -1.480349021      0
#> 322   1.257344964      0
#> 323   0.971423603      0
#> 324   0.588072527      0
#> 325  -2.073105917      0
#> 326  -0.939433871      0
#> 327   1.340942825      0
#> 328   1.640179928      0
#> 329   0.275890131      0
#> 330  -1.986181608      0
#> 331  -0.700581614      0
#> 332   0.742173097      0
#> 333  -0.326680194      0
#> 334  -0.533541723      0
#> 335  -0.545073196      0
#> 336  -1.638208946      0
#> 337  -0.725084412      0
#> 338   0.154013129      0
#> 339   1.349437157      0
#> 340  -0.535360755      0
#> 341  -0.459894457      0
#> 342   0.600934141      0
#> 343  -2.014189996      0
#> 344  -0.440494282      0
#> 345   2.478566351      0
#> 346   1.515071390      0
#> 347  -0.125653019      0
#> 348   1.253095472      0
#> 349  -0.478133644      0
#> 350   1.143150648      0
#> 351   0.201891370      0
#> 352   1.376452553      0
#> 353  -1.064120960      0
#> 354   0.629747939      0
#> 355  -2.202594546      0
#> 356   0.067922251      0
#> 357  -1.291363943      0
#> 358  -2.430515984      0
#> 359  -1.202919131      0
#> 360   0.624490266      0
#> 361   0.059924042      0
#> 362   1.937340634      0
#> 363   0.095666944      0
#> 364   0.626665445      0
#> 365  -0.716799677      0
#> 366   0.511714830      0
#> 367   0.837981454      0
#> 368   0.734084403      0
#> 369  -1.239059443      0
#> 370   0.922891275      0
#> 371  -0.256291784      0
#> 372   1.078564899      0
#> 373  -2.613990459      0
#> 374  -1.638096000      0
#> 375   1.306242521      0
#> 376   1.035936379      0
#> 377  -0.504463677      0
#> 378  -0.686428807      0
#> 379  -0.417136651      0
#> 380  -1.474463149      0
#> 381   0.266726721      0
#> 382   0.387003663      0
#> 383  -1.049670422      0
#> 384   0.142016845      0
#> 385   2.148662202      0
#> 386   0.214224438      0
#> 387  -0.857001454      0
#> 388  -0.740139936      0
#> 389  -1.298930985      0
#> 390   0.031822476      0
#> 391  -0.941838685      0
#> 392  -0.020526310      0
#> 393  -0.919435565      0
#> 394  -0.413384310      0
#> 395  -0.506693454      0
#> 396  -0.665960779      0
#> 397  -1.025261049      0
#> 398   0.380931890      0
#> 399  -0.719847576      0
#> 400  -0.131717312      0
#> 401   0.683055554      0
#> 402   2.350550515      0
#> 403  -0.556016249      0
#> 404   1.686263136      0
#> 405  -0.149898525      0
#> 406  -1.038689700      0
#> 407   1.794911592      0
#> 408   1.326994203      0
#> 409  -1.618593395      0
#> 410  -1.893725114      0
#> 411   0.575406038      0
#> 412  -0.777801424      0
#> 413  -0.078677298      0
#> 414  -0.709712137      0
#> 415   1.643273856      0
#> 416   1.156389215      0
#> 417   0.581726296      0
#> 418  -1.778093784      0
#> 419  -1.497241887      0
#> 420   0.554208760      0
#> 421   0.342903688      0
#> 422  -1.203041256      0
#> 423  -0.446103861      0
#> 424   2.020249622      0
#> 425  -0.771821549      0
#> 426   2.063570849      0
#> 427  -0.673559094      0
#> 428   0.019352330      0
#> 429   0.703265184      0
#> 430  -1.024161663      0
#> 431   1.740306113      0
#> 432   0.084183911      0
#> 433   0.005964424      0
#> 434  -0.422410941      0
#> 435  -1.996066133      0
#> 436   0.038755737      0
#> 437  -1.304655088      0
#> 438  -0.837456801      0
#> 439   1.525892198      0
#> 440  -1.646426713      0
#> 441  -1.518521563      0
#> 442   1.362908728      0
#> 443   1.778493278      0
#> 444   0.591574746      0
#> 445  -0.220398418      0
#> 446   1.768567605      0
#> 447   0.224119809      0
#> 448   0.351986364      0
#> 449   0.813284176      0
#> 450  -1.052184035      0
#> 451  -0.776291669      0
#> 452  -0.177131714      0
#> 453  -2.058400212      0
#> 454  -0.469768415      0
#> 455  -1.291873454      0
#> 456   0.774752776      0
#> 457  -1.437569320      0
#> 458  -0.329301178      0
#> 459   1.499429599      0
#> 460   2.839401557      0
#> 461   0.329689833      0
#> 462   1.851325969      0
#> 463   0.375818560      0
#> 464   1.117744013      0
#> 465  -3.523729360      0
#> 466   0.405642867      0
#> 467   0.818065713      0
#> 468   1.483057966      0
#> 469  -1.086362995      0
#> 470   1.502918143      0
#> 471   0.321688770      0
#> 472  -1.002718483      0
#> 473   0.872875487      0
#> 474   0.448818486      0
#> 475   0.685828211      0
#> 476  -0.290953052      0
#> 477   0.404334793      0
#> 478  -1.178652155      0
#> 479   0.280120217      0
#> 480   1.893984042      0
#> 481   1.181971253      0
#> 482  -0.998648934      0
#> 483   1.125953773      0
#> 484  -0.637536219      0
#> 485   0.583516783      0
#> 486  -0.703189855      0
#> 487  -0.517325104      0
#> 488  -0.954713092      0
#> 489  -0.899692599      0
#> 490  -0.836464994      0
#> 491   0.465902507      0
#> 492   1.166275619      0
#> 493   0.631238058      0
#> 494   0.849412800      0
#> 495  -0.797495636      0
#> 496   2.083815301      0
#> 497  -0.265143389      0
#> 498   0.481912908      0
#> 499   0.378672341      0
#> 500   2.285625395      0
#> 501  -0.652058626      0
#> 502  -1.673144526      0
#> 503   1.669088524      0
#> 504   0.800363156      0
#> 505   1.442098686      0
#> 506  -0.454708406      0
#> 507  -0.997219885      0
#> 508   1.190443781      0
#> 509   1.252636145      0
#> 510  -1.514702744      0
#> 511   0.420209476      0
#> 512  -1.328938707      0
#> 513  -0.996191382      0
#> 514  -0.054083830      0
#> 515   0.400389258      0
#> 516   1.002748819      0
#> 517   1.897101991      0
#> 518  -0.235784029      0
#> 519  -1.066591404      0
#> 520  -0.451835776      0
#> 521   0.001978982      0
#> 522   0.070802775      0
#> 523  -0.149449041      0
#> 524  -0.434181731      0
#> 525   0.434323858      0
#> 526  -0.290322451      0
#> 527  -0.904611463      0
#> 528   0.795305413      0
#> 529  -0.558273862      0
#> 530   0.140656731      0
#> 531   0.252790053      0
#> 532   0.062765206      0
#> 533  -2.669753764      0
#> 534   1.107575683      0
#> 535   1.664418560      0
#> 536  -0.552913149      0
#> 537  -0.430074045      0
#> 538   0.432548710      0
#> 539  -0.149765079      0
#> 540   0.101351703      0
#> 541  -0.815324448      0
#> 542  -0.559242650      0
#> 543   1.677748698      0
#> 544  -1.329323119      0
#> 545  -1.749989523      0
#> 546   0.712827423      0
#> 547   0.822640361      0
#> 548   0.903742822      0
#> 549  -0.436541065      0
#> 550   0.383190593      0
#> 551   0.744528446      0
#> 552   1.215194006      0
#> 553   0.194060396      0
#> 554  -1.770571339      0
#> 555   1.193679841      0
#> 556  -1.903569294      0
#> 557   0.899285603      0
#> 558   0.165115850      0
#> 559  -0.696550733      0
#> 560  -0.337779863      0
#> 561   0.588479206      0
#> 562  -0.016974933      0
#> 563   0.958050379      0
#> 564   0.479679861      0
#> 565   1.074054082      0
#> 566  -1.281544634      0
#> 567  -0.540359631      0
#> 568  -0.318099458      0
#> 569   0.789204403      0
#> 570   0.363321546      0
#> 571  -0.221651015      0
#> 572  -1.202707622      0
#> 573   0.882632663      0
#> 574   0.879804843      0
#> 575  -1.685600153      0
#> 576  -2.068366467      0
#> 577   0.744380726      0
#> 578  -0.617838091      0
#> 579  -0.651804984      0
#> 580  -1.072186515      0
#> 581  -1.609664899      0
#> 582   0.695463218      0
#> 583   0.422638525      0
#> 584  -0.238915878      0
#> 585   1.454507244      0
#> 586  -1.612709730      0
#> 587   0.691212212      0
#> 588   0.455474048      0
#> 589  -0.892055672      0
#> 590  -0.526388584      0
#> 591   1.089744836      0
#> 592  -2.373110586      0
#> 593   0.876399255      0
#> 594   0.229706584      0
#> 595  -0.534821473      0
#> 596  -0.444749342      0
#> 597  -0.794922952      0
#> 598  -1.475457868      0
#> 599   0.674264473      0
#> 600  -0.094276939      0
#> 601   0.690287620      0
#> 602  -0.891069790      0
#> 603  -0.055726713      0
#> 604  -0.159245661      0
#> 605  -1.201048687      0
#> 606   0.274400327      0
#> 607   0.210032773      0
#> 608   0.982912016      0
#> 609   0.387724037      0
#> 610   1.427619681      0
#> 611   0.202413118      0
#> 612  -0.483298290      0
#> 613  -0.255185617      0
#> 614   0.702102426      0
#> 615   0.163576022      0
#> 616  -0.916390772      0
#> 617   2.258619064      0
#> 618   0.819185972      0
#> 619  -1.131131988      0
#> 620   0.098515478      0
#> 621   0.382161780      0
#> 622   0.885194330      0
#> 623   0.994606384      0
#> 624   0.563328929      0
#> 625   0.753421942      0
#> 626  -0.419616116      0
#> 627  -0.618428411      0
#> 628   0.397018225      0
#> 629  -0.559558103      0
#> 630   1.441083258      0
#> 631   0.256938412      0
#> 632   0.758511904      0
#> 633   0.245584729      0
#> 634  -0.281596135      0
#> 635  -1.062143688      0
#> 636   2.189893032      0
#> 637   1.205895606      0
#> 638   0.649585950      0
#> 639  -1.509976936      0
#> 640  -0.061470639      0
#> 641  -0.647423257      0
#> 642   2.284408465      0
#> 643  -1.327669812      0
#> 644  -1.919350827      0
#> 645   1.476895304      0
#> 646  -0.562959312      0
#> 647   1.461941448      0
#> 648  -0.630234589      0
#> 649  -0.058198470      0
#> 650  -0.376268523      0
#> 651  -0.679148724      0
#> 652  -0.273061758      0
#> 653  -0.413112908      0
#> 654   0.974660219      0
#> 655  -0.380080239      0
#> 656   1.300248771      0
#> 657   0.248742602      0
#> 658  -1.389774669      0
#> 659   0.625373765      0
#> 660  -0.931501185      0
#> 661  -0.321577107      0
#> 662  -1.561541145      0
#> 663   0.056347316      0
#> 664   0.352589440      0
#> 665  -1.971360133      0
#> 666  -0.023827552      0
#> 667  -0.564426704      0
#> 668  -0.279157044      0
#> 669   0.271976282      0
#> 670   0.664003048      0
#> 671  -1.243525189      0
#> 672  -1.105355709      0
#> 673  -0.262150108      0
#> 674  -0.192821731      0
#> 675   0.401464659      0
#> 676  -1.284043608      0
#> 677  -1.085170308      0
#> 678  -1.955145457      0
#> 679  -0.794161628      0
#> 680  -0.260402876      0
#> 681  -1.340550062      0
#> 682   0.634961584      0
#> 683  -1.479251159      0
#> 684  -0.927194570      0
#> 685  -0.485001155      0
#> 686  -0.419187460      0
#> 687  -1.176753875      0
#> 688  -1.149025363      0
#> 689   0.193283193      0
#> 690   0.681931601      0
#> 691   1.457992666      0
#> 692  -0.390057640      0
#> 693  -0.325309508      0
#> 694   1.626405578      0
#> 695  -0.592961918      0
#> 696  -0.494969867      0
#> 697  -1.624789007      0
#> 698   0.303893103      0
#> 699   1.263021536      0
#> 700   1.034116597      0
#> 701  -0.085959149      0
#> 702  -0.557672680      0
#> 703   0.018873677      0
#> 704  -0.258496436      0
#> 705   1.068227441      0
#> 706   0.241346995      0
#> 707  -0.755264826      0
#> 708  -1.393059901      0
#> 709   0.017834019      0
#> 710   1.022044655      0
#> 711  -2.441144848      0
#> 712  -0.759070872      0
#> 713  -1.499530964      0
#> 714  -1.364117525      0
#> 715   0.479748126      0
#> 716   0.165851657      0
#> 717   2.066688469      0
#> 718   0.219269602      0
#> 719   0.002706847      0
#> 720   2.234965017      0
#> 721  -0.253671224      0
#> 722   0.779151050      0
#> 723   0.242820986      0
#> 724  -1.366214434      0
#> 725  -0.283067005      0
#> 726   0.482099218      0
#> 727   0.783214912      0
#> 728   0.840498191      0
#> 729   0.508779796      0
#> 730  -1.675700119      0
#> 731   0.207516065      0
#> 732   0.104566292      0
#> 733  -0.131696236      0
#> 734  -0.680524805      0
#> 735  -0.539506436      0
#> 736  -0.269848284      0
#> 737   0.187935524      0
#> 738  -0.631496018      0
#> 739   1.180627072      0
#> 740  -2.184461405      0
#> 741  -0.761821578      0
#> 742  -1.622734885      0
#> 743   1.474697459      0
#> 744   1.346670951      0
#> 745  -0.914433449      0
#> 746  -0.956737601      0
#> 747  -1.759625187      0
#> 748  -0.557327710      0
#> 749   1.486423187      0
#> 750  -0.151621849      0
#> 751   0.214062778      0
#> 752  -1.451145270      0
#> 753  -1.294316879      0
#> 754   1.001400813      0
#> 755  -0.631459940      0
#> 756  -1.130465423      0
#> 757   1.518230293      0
#> 758   0.743249935      0
#> 759  -0.096169038      0
#> 760  -1.094596659      0
#> 761  -0.509462069      0
#> 762  -0.572787810      0
#> 763   0.399653686      0
#> 764   0.050968099      0
#> 765   0.839868324      0
#> 766  -0.093069826      0
#> 767  -0.127547815      0
#> 768  -0.295487122      0
#> 769   0.258864182      0
#> 770  -1.058797969      0
#> 771  -0.143867763      0
#> 772   0.239724741      0
#> 773  -0.249179173      0
#> 774   0.082874270      0
#> 775   1.749310270      0
#> 776  -0.917859855      0
#> 777  -0.332199859      0
#> 778   2.006112731      0
#> 779   2.056942653      0
#> 780   1.652807936      0
#> 781   0.280659823      0
#> 782   0.512106806      0
#> 783   0.195819983      0
#> 784  -0.585915221      0
#> 785  -0.739683126      0
#> 786   1.371247386      0
#> 787   0.715057500      0
#> 788   0.961128893      0
#> 789  -0.564325149      0
#> 790   1.001061045      0
#> 791   0.546147860      0
#> 792   0.510093697      0
#> 793  -0.345017014      0
#> 794   0.570775688      0
#> 795  -0.054381184      0
#> 796  -0.287960786      0
#> 797   0.834986818      0
#> 798  -1.152158449      0
#> 799  -2.094438922      0
#> 800  -1.122698584      0
#> 801   0.354002730      0
#> 802   0.161338416      0
#> 803  -0.205530460      0
#> 804   1.509572084      0
#> 805  -0.609677537      0
#> 806   0.048483545      0
#> 807   2.829784509      0
#> 808  -0.239654582      0
#> 809   0.162078225      0
#> 810   1.652967940      0
#> 811   0.383502099      0
#> 812   0.626622520      0
#> 813   0.184299115      0
#> 814   0.547238055      0
#> 815  -0.198560507      0
#> 816   0.657607831      0
#> 817  -1.583369962      0
#> 818  -0.450834943      0
#> 819  -1.161079124      0
#> 820  -0.541430747      0
#> 821  -0.458882164      0
#> 822  -1.126773467      0
#> 823  -0.980685289      0
#> 824  -0.250323982      0
#> 825   0.597248577      0
#> 826  -0.583806288      0
#> 827  -0.456627139      0
#> 828  -0.022511172      0
#> 829   1.607544920      0
#> 830  -0.349084442      0
#> 831   0.204463602      0
#> 832   0.319716661      0
#> 833  -0.153837276      0
#> 834   0.499416958      0
#> 835   0.528688745      0
#> 836  -0.288317359      0
#> 837   0.192023826      0
#> 838  -0.986554975      0
#> 839  -0.295690042      0
#> 840  -0.390662594      0
#> 841   0.003527854      0
#> 842  -1.136538294      0
#> 843  -2.153072977      0
#> 844  -0.200158782      0
#> 845   0.346062864      0
#> 846   0.168101216      0
#> 847   0.350410102      0
#> 848   0.631394315      0
#> 849   1.371537032      0
#> 850  -0.777873482      0
#> 851   1.073301189      0
#> 852  -1.002120738      0
#> 853  -0.459977162      0
#> 854  -0.307444748      0
#> 855  -1.145792165      0
#> 856  -0.624736669      0
#> 857  -0.875150523      0
#> 858   0.495965934      0
#> 859  -1.297107609      0
#> 860   1.039876000      0
#> 861   0.304578992      0
#> 862   2.605190157      0
#> 863  -0.796631493      0
#> 864   1.612555219      0
#> 865   0.719162093      0
#> 866   1.054743679      0
#> 867   0.029431151      0
#> 868   1.205787645      0
#> 869  -1.983584412      0
#> 870  -0.776843978      0
#> 871   0.122242009      0
#> 872  -0.291093396      0
#> 873   0.174205833      0
#> 874   1.746835949      0
#> 875   0.370631059      0
#> 876   1.233935968      0
#> 877  -0.208640953      0
#> 878   0.639591232      0
#> 879   1.429719405      0
#> 880   1.312734189      0
#> 881  -0.162616038      0
#> 882  -0.011461406      0
#> 883  -0.423270573      0
#> 884   1.098905599      0
#> 885  -0.562925099      0
#> 886   0.180921849      0
#> 887  -0.889864831      0
#> 888   0.509716499      0
#> 889  -0.279406928      0
#> 890   1.741337519      0
#> 891  -0.763417032      0
#> 892   1.258803807      0
#> 893   0.613121329      0
#> 894   2.005976048      0
#> 895   1.096798570      0
#> 896   0.420034260      0
#> 897  -0.704725367      0
#> 898  -1.355781285      0
#> 899   0.467530036      0
#> 900   0.560460691      0
#> 901   1.369434944      1
#> 902   1.008531630      1
#> 903  -0.284853282      1
#> 904  -0.096747132      1
#> 905   2.458664616      1
#> 906  -0.274659358      1
#> 907   1.394782246      1
#> 908   2.015182322      1
#> 909  -0.069568994      1
#> 910   0.791184032      1
#> 911   2.709821491      1
#> 912   2.181458989      1
#> 913  -0.315162890      1
#> 914   1.170035640      1
#> 915  -0.083628546      1
#> 916   2.465400902      1
#> 917   0.592149003      1
#> 918  -0.026300089      1
#> 919   0.401050886      1
#> 920   3.132733328      1
#> 921  -0.379231745      1
#> 922   0.824383423      1
#> 923  -1.332984293      1
#> 924   0.690895529      1
#> 925  -0.023430782      1
#> 926   1.843938030      1
#> 927  -0.114858885      1
#> 928   0.332721935      1
#> 929   0.050583742      1
#> 930   0.645779936      1
#> 931   1.738696497      1
#> 932   1.160149417      1
#> 933   1.356472198      1
#> 934   0.681973659      1
#> 935   0.794001516      1
#> 936   0.725639613      1
#> 937   1.954082510      1
#> 938   1.155796148      1
#> 939   1.164539032      1
#> 940   2.013794771      1
#> 941   2.094463827      1
#> 942  -0.080423383      1
#> 943   0.574020463      1
#> 944  -0.896165205      1
#> 945  -0.036972704      1
#> 946  -0.229101560      1
#> 947   1.476498121      1
#> 948   1.834671421      1
#> 949  -1.018226422      1
#> 950   0.865785017      1
#> 951   2.102647708      1
#> 952   1.308291171      1
#> 953   0.206835492      1
#> 954  -0.168696993      1
#> 955   1.371386653      1
#> 956   0.047350532      1
#> 957   0.881229207      1
#> 958   0.625948822      1
#> 959   0.714485545      1
#> 960  -0.403757024      1
#> 961   0.223211148      1
#> 962   1.356015518      1
#> 963  -0.562664906      1
#> 964   1.174103371      1
#> 965   0.175578921      1
#> 966  -0.088032907      1
#> 967   0.111453235      1
#> 968  -0.744840450      1
#> 969  -0.414112001      1
#> 970   1.221285376      1
#> 971   2.031936315      1
#> 972  -1.123261550      1
#> 973  -0.456669202      1
#> 974   1.568873772      1
#> 975   1.082000656      1
#> 976   1.289679341      1
#> 977   0.590477198      1
#> 978   1.323143462      1
#> 979   1.611654546      1
#> 980   0.492133111      1
#> 981   2.163670998      1
#> 982   0.239818462      1
#> 983   0.681039071      1
#> 984   1.109285470      1
#> 985   0.146106046      1
#> 986   0.008568352      1
#> 987   1.430254544      1
#> 988   1.392550012      1
#> 989  -0.362549294      1
#> 990   1.263436891      1
#> 991   0.381533653      1
#> 992   0.458951094      1
#> 993   0.796516680      1
#> 994   0.179475552      1
#> 995   3.478924291      1
#> 996   2.415507603      1
#> 997   2.341555119      1
#> 998   0.627116277      1
#> 999   2.490052166      1
#> 1000 -0.702899864      1