Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     1.020028e+00      0
#> 2     1.304269e+00      0
#> 3    -1.821391e+00      0
#> 4    -2.702660e-01      0
#> 5     8.818399e-01      0
#> 6     3.417944e-01      0
#> 7    -2.402127e-01      0
#> 8    -1.236051e+00      0
#> 9     1.134183e+00      0
#> 10    1.922170e-01      0
#> 11    1.223060e+00      0
#> 12    1.693108e+00      0
#> 13    7.591383e-01      0
#> 14    1.855512e+00      0
#> 15   -5.246685e-01      0
#> 16    8.328397e-02      0
#> 17    1.025953e+00      0
#> 18    1.363709e+00      0
#> 19   -1.052284e-01      0
#> 20   -1.480912e-01      0
#> 21    7.053884e-01      0
#> 22    3.772867e-03      0
#> 23    1.049260e-01      0
#> 24    3.437208e-01      0
#> 25   -2.588908e+00      0
#> 26    1.387114e+00      0
#> 27   -7.849710e-01      0
#> 28   -3.049073e-01      0
#> 29    1.276318e+00      0
#> 30    2.326459e-01      0
#> 31   -5.747464e-01      0
#> 32   -1.729971e+00      0
#> 33   -3.783853e-01      0
#> 34    4.468646e-01      0
#> 35   -2.850808e-01      0
#> 36    1.722024e+00      0
#> 37   -1.456599e+00      0
#> 38    1.280645e-01      0
#> 39    7.208385e-01      0
#> 40   -6.696547e-01      0
#> 41   -1.479425e+00      0
#> 42   -1.602848e-01      0
#> 43    4.691201e-01      0
#> 44    6.412681e-01      0
#> 45    1.151456e-01      0
#> 46   -2.286575e+00      0
#> 47    4.264983e-01      0
#> 48    6.274597e-01      0
#> 49   -9.429046e-01      0
#> 50   -5.416635e-01      0
#> 51    1.362988e+00      0
#> 52   -3.962694e-02      0
#> 53   -6.818576e-01      0
#> 54   -4.472456e-01      0
#> 55   -7.336496e-01      0
#> 56   -4.240443e-01      0
#> 57   -2.547245e-01      0
#> 58   -3.608393e-01      0
#> 59    1.866736e-03      0
#> 60   -9.222273e-01      0
#> 61   -5.185299e-01      0
#> 62    9.712866e-01      0
#> 63   -1.024432e+00      0
#> 64   -4.156998e-01      0
#> 65    1.372455e+00      0
#> 66   -1.881863e-01      0
#> 67   -1.491958e-01      0
#> 68    1.545662e+00      0
#> 69    1.301594e+00      0
#> 70    2.764759e-01      0
#> 71    1.135929e+00      0
#> 72   -2.971469e-01      0
#> 73   -8.154239e-02      0
#> 74    2.334590e-01      0
#> 75    1.131495e+00      0
#> 76    1.652400e+00      0
#> 77    6.067640e-01      0
#> 78   -5.718703e-01      0
#> 79   -1.355846e+00      0
#> 80   -5.554812e-01      0
#> 81    7.230622e-01      0
#> 82    2.044767e+00      0
#> 83    6.408067e-01      0
#> 84   -7.765803e-01      0
#> 85   -1.561879e+00      0
#> 86    9.290257e-01      0
#> 87   -3.759425e-01      0
#> 88   -1.229590e+00      0
#> 89   -1.063741e+00      0
#> 90   -1.079063e-01      0
#> 91   -1.156371e-01      0
#> 92   -1.046834e+00      0
#> 93   -1.464499e+00      0
#> 94   -4.529229e-01      0
#> 95    1.237645e+00      0
#> 96   -7.336600e-02      0
#> 97   -6.882556e-01      0
#> 98    1.206150e+00      0
#> 99    1.167229e+00      0
#> 100  -7.180752e-01      0
#> 101   1.356039e-01      0
#> 102  -7.373178e-01      0
#> 103   4.686516e-02      0
#> 104  -3.555459e-01      0
#> 105   9.031707e-02      0
#> 106   1.103327e+00      0
#> 107  -1.033171e+00      0
#> 108  -4.100122e-01      0
#> 109  -1.415593e+00      0
#> 110   9.573032e-01      0
#> 111  -6.442585e-01      0
#> 112  -2.893523e-01      0
#> 113   5.574005e-01      0
#> 114  -1.512347e-01      0
#> 115   3.449981e-01      0
#> 116  -1.698144e+00      0
#> 117  -1.562907e+00      0
#> 118   7.353699e-02      0
#> 119  -2.966224e-01      0
#> 120  -1.758720e-01      0
#> 121  -8.844565e-01      0
#> 122  -1.745260e+00      0
#> 123  -1.706891e+00      0
#> 124  -6.269327e-01      0
#> 125   8.654596e-01      0
#> 126  -1.195665e+00      0
#> 127   1.690200e-01      0
#> 128  -1.040512e-01      0
#> 129   2.732675e+00      0
#> 130  -8.547149e-01      0
#> 131  -1.158310e+00      0
#> 132   1.574751e+00      0
#> 133  -2.281941e-01      0
#> 134   1.619302e+00      0
#> 135  -9.995279e-01      0
#> 136   4.039814e-01      0
#> 137  -6.125724e-01      0
#> 138  -7.625280e-01      0
#> 139   4.199979e-02      0
#> 140   9.924916e-01      0
#> 141  -2.927100e+00      0
#> 142   7.305456e-01      0
#> 143   5.250737e-01      0
#> 144   4.170946e-01      0
#> 145  -7.615073e-01      0
#> 146   3.397239e-01      0
#> 147   1.364428e+00      0
#> 148   2.243526e-01      0
#> 149   1.008556e+00      0
#> 150  -1.066117e+00      0
#> 151   1.532617e+00      0
#> 152   1.133667e+00      0
#> 153   8.927767e-01      0
#> 154  -7.634262e-01      0
#> 155   5.159066e-01      0
#> 156  -1.295115e-01      0
#> 157  -2.098810e+00      0
#> 158  -3.746313e-02      0
#> 159   3.484619e-01      0
#> 160  -1.035290e+00      0
#> 161   1.383254e+00      0
#> 162  -1.454556e+00      0
#> 163  -2.139202e-01      0
#> 164   3.471923e-02      0
#> 165   5.337071e-01      0
#> 166   7.336038e-01      0
#> 167  -5.288901e-01      0
#> 168  -8.083984e-01      0
#> 169  -3.783901e-01      0
#> 170  -1.504336e+00      0
#> 171  -1.114941e+00      0
#> 172  -9.531464e-01      0
#> 173  -2.456548e-01      0
#> 174   3.300081e-01      0
#> 175  -7.000830e-01      0
#> 176   1.331991e+00      0
#> 177  -4.058954e-01      0
#> 178  -1.446388e+00      0
#> 179  -7.815565e-01      0
#> 180  -4.259837e-05      0
#> 181   8.942756e-01      0
#> 182   7.510459e-01      0
#> 183  -3.787708e-01      0
#> 184   1.731441e-01      0
#> 185  -9.283589e-01      0
#> 186   1.606406e+00      0
#> 187   4.703439e-01      0
#> 188   2.945873e-01      0
#> 189   1.109141e+00      0
#> 190   1.043168e+00      0
#> 191  -4.999114e-02      0
#> 192   6.251862e-01      0
#> 193  -2.654678e-01      0
#> 194   2.109095e-01      0
#> 195   1.483048e-01      0
#> 196  -4.975428e-03      0
#> 197  -4.086212e-01      0
#> 198  -1.201089e+00      0
#> 199  -4.106472e-01      0
#> 200   2.496629e-01      0
#> 201   4.334097e-02      0
#> 202  -1.039676e+00      0
#> 203  -1.054173e-02      0
#> 204  -1.070800e+00      0
#> 205  -2.893862e+00      0
#> 206   6.150319e-01      0
#> 207  -6.579438e-01      0
#> 208  -5.573496e-01      0
#> 209   1.938212e+00      0
#> 210  -6.231044e-01      0
#> 211  -4.100013e-02      0
#> 212  -1.742990e+00      0
#> 213   5.860067e-01      0
#> 214   8.333006e-01      0
#> 215  -6.280450e-01      0
#> 216  -1.708163e+00      0
#> 217   9.014050e-01      0
#> 218  -1.987108e+00      0
#> 219   1.552961e-01      0
#> 220   5.294533e-01      0
#> 221   1.095570e-01      0
#> 222  -1.096850e+00      0
#> 223  -1.770306e-01      0
#> 224   8.934272e-01      0
#> 225   1.191597e-01      0
#> 226  -6.152146e-01      0
#> 227   4.171982e-01      0
#> 228   3.396511e-02      0
#> 229  -7.535569e-02      0
#> 230   6.068635e-01      0
#> 231   1.226065e-01      0
#> 232  -4.339348e-02      0
#> 233   1.788424e+00      0
#> 234  -2.372478e+00      0
#> 235  -2.817331e+00      0
#> 236  -2.966349e-01      0
#> 237  -1.335966e+00      0
#> 238   3.082792e-01      0
#> 239  -1.238357e+00      0
#> 240  -3.351366e-02      0
#> 241   1.373868e+00      0
#> 242   2.411270e-01      0
#> 243  -7.648591e-01      0
#> 244   3.041349e-01      0
#> 245  -2.089552e-01      0
#> 246   2.242215e-01      0
#> 247   7.575786e-01      0
#> 248   4.526941e-01      0
#> 249   4.754499e-01      0
#> 250   4.198393e-01      0
#> 251   1.252687e+00      0
#> 252  -1.448686e-01      0
#> 253   1.246099e+00      0
#> 254   9.182553e-01      0
#> 255  -1.984077e+00      0
#> 256   7.100392e-01      0
#> 257   6.319043e-02      0
#> 258  -1.288219e+00      0
#> 259   6.523698e-01      0
#> 260  -7.898136e-01      0
#> 261  -1.625746e+00      0
#> 262  -5.268123e-01      0
#> 263  -3.135931e-01      0
#> 264   5.459303e-01      0
#> 265   1.403662e+00      0
#> 266  -1.732733e+00      0
#> 267   1.885075e+00      0
#> 268   3.089076e-01      0
#> 269  -1.585957e-01      0
#> 270   9.518374e-01      0
#> 271   2.876812e+00      0
#> 272   5.242331e-01      0
#> 273  -2.814686e-01      0
#> 274  -3.223859e-01      0
#> 275  -1.646967e-01      0
#> 276  -9.354748e-01      0
#> 277   1.039608e-01      0
#> 278   1.549538e+00      0
#> 279   2.836864e-01      0
#> 280  -1.286084e+00      0
#> 281   2.762996e-01      0
#> 282   3.964369e-01      0
#> 283  -1.208238e+00      0
#> 284   3.234991e-01      0
#> 285  -1.141071e+00      0
#> 286  -1.062634e+00      0
#> 287   1.543640e+00      0
#> 288   2.853195e-02      0
#> 289   1.756732e+00      0
#> 290  -1.879352e+00      0
#> 291   2.522441e-01      0
#> 292   9.249973e-01      0
#> 293  -1.591302e+00      0
#> 294  -3.519891e-01      0
#> 295   2.040244e+00      0
#> 296   4.457941e-01      0
#> 297  -4.664925e-01      0
#> 298   7.455372e-01      0
#> 299   1.519741e+00      0
#> 300  -5.378412e-01      0
#> 301  -2.878330e-01      0
#> 302   1.494648e+00      0
#> 303   2.030913e-02      0
#> 304   1.115385e+00      0
#> 305  -7.924712e-01      0
#> 306  -6.442447e-01      0
#> 307   9.223490e-01      0
#> 308  -2.032924e-01      0
#> 309  -3.000360e+00      0
#> 310   1.148544e+00      0
#> 311   8.605316e-02      0
#> 312  -1.009185e+00      0
#> 313   1.612548e-01      0
#> 314  -6.402330e-01      0
#> 315   4.599630e-01      0
#> 316   2.547835e+00      0
#> 317   8.184516e-01      0
#> 318  -5.364045e-01      0
#> 319   1.935058e+00      0
#> 320   1.691540e+00      0
#> 321  -2.181887e-01      0
#> 322   1.963473e-01      0
#> 323  -9.845455e-01      0
#> 324   7.870357e-01      0
#> 325   9.330787e-01      0
#> 326  -1.093893e+00      0
#> 327  -1.999094e+00      0
#> 328   2.662475e-01      0
#> 329  -1.230555e+00      0
#> 330  -1.239163e+00      0
#> 331  -5.072518e-01      0
#> 332  -3.908041e-01      0
#> 333   6.390682e-01      0
#> 334  -6.586422e-01      0
#> 335   2.171566e-01      0
#> 336  -1.031842e+00      0
#> 337  -7.192750e-01      0
#> 338  -4.626275e-01      0
#> 339  -1.251084e+00      0
#> 340   9.827017e-01      0
#> 341  -5.385986e-01      0
#> 342   1.544222e+00      0
#> 343   1.033516e+00      0
#> 344   4.062477e-01      0
#> 345   1.571637e-01      0
#> 346  -8.216262e-01      0
#> 347   3.642086e-01      0
#> 348   6.010922e-01      0
#> 349  -7.100026e-01      0
#> 350  -4.657000e-01      0
#> 351   9.978845e-02      0
#> 352   1.219957e+00      0
#> 353  -2.178907e-01      0
#> 354  -2.605056e-02      0
#> 355  -1.048933e+00      0
#> 356  -3.504991e-01      0
#> 357   5.118085e-01      0
#> 358  -1.381857e+00      0
#> 359  -5.061463e-01      0
#> 360  -4.545420e-01      0
#> 361   1.026473e+00      0
#> 362   4.950763e-01      0
#> 363   1.642123e+00      0
#> 364  -8.112518e-01      0
#> 365   4.134023e-01      0
#> 366  -3.526532e-01      0
#> 367   1.410232e+00      0
#> 368   5.873881e-01      0
#> 369   5.508691e-01      0
#> 370  -1.009898e+00      0
#> 371   5.839066e-01      0
#> 372  -7.985905e-01      0
#> 373   8.077620e-02      0
#> 374   7.142008e-01      0
#> 375   2.207186e-01      0
#> 376  -7.860312e-02      0
#> 377  -1.877914e+00      0
#> 378  -9.092039e-01      0
#> 379  -2.158903e-01      0
#> 380  -9.777186e-01      0
#> 381  -9.149267e-01      0
#> 382   2.393501e+00      0
#> 383   8.325396e-01      0
#> 384   1.202151e+00      0
#> 385  -2.231976e-01      0
#> 386  -1.592525e+00      0
#> 387   7.340009e-01      0
#> 388  -6.186347e-01      0
#> 389  -5.258303e-01      0
#> 390  -1.604101e+00      0
#> 391   3.266595e-01      0
#> 392   1.466955e+00      0
#> 393   3.001768e-01      0
#> 394   1.954682e-01      0
#> 395  -3.425399e-01      0
#> 396  -8.124145e-01      0
#> 397  -9.435954e-01      0
#> 398   1.242934e+00      0
#> 399   8.865394e-01      0
#> 400  -2.219329e-01      0
#> 401   1.180716e+00      0
#> 402  -1.515302e+00      0
#> 403  -6.489507e-01      0
#> 404  -9.187792e-01      0
#> 405   1.863269e-01      0
#> 406   7.741674e-01      0
#> 407   6.356844e-03      0
#> 408  -3.625180e-01      0
#> 409  -1.309329e+00      0
#> 410   6.117462e-01      0
#> 411   5.112304e-02      0
#> 412  -5.222070e-01      0
#> 413   1.331690e+00      0
#> 414   1.101850e-01      0
#> 415   1.925061e-01      0
#> 416   7.360285e-01      0
#> 417   2.648550e-01      0
#> 418   1.372500e+00      0
#> 419   7.591152e-01      0
#> 420  -9.103327e-02      0
#> 421   2.866738e-01      0
#> 422  -1.373548e+00      0
#> 423   2.017405e+00      0
#> 424   1.949027e-01      0
#> 425   6.378081e-01      0
#> 426   7.900511e-01      0
#> 427   5.472042e-01      0
#> 428   8.765840e-01      0
#> 429   9.872508e-01      0
#> 430  -3.478966e-02      0
#> 431   2.284545e+00      0
#> 432   2.378092e-01      0
#> 433   4.539702e-01      0
#> 434   5.170610e-02      0
#> 435   2.943843e-01      0
#> 436   8.215202e-01      0
#> 437   1.292118e+00      0
#> 438   1.578871e-01      0
#> 439  -6.301719e-01      0
#> 440  -1.089426e-01      0
#> 441   9.087780e-01      0
#> 442   5.220345e-01      0
#> 443  -9.533545e-01      0
#> 444  -5.355286e-01      0
#> 445  -2.504787e+00      0
#> 446   3.978110e-01      0
#> 447  -1.218778e+00      0
#> 448   3.667239e-01      0
#> 449   1.698787e+00      0
#> 450  -1.483146e+00      0
#> 451   4.954427e-01      0
#> 452  -1.244330e-01      0
#> 453   1.181877e+00      0
#> 454   2.936070e-01      0
#> 455  -8.942264e-01      0
#> 456  -1.116982e+00      0
#> 457  -1.309510e+00      0
#> 458  -1.009579e+00      0
#> 459   1.379561e+00      0
#> 460  -1.314008e+00      0
#> 461  -3.502628e-01      0
#> 462   1.658636e+00      0
#> 463  -1.668816e+00      0
#> 464  -2.974297e-01      0
#> 465   9.035911e-01      0
#> 466  -5.709689e-01      0
#> 467   1.308612e+00      0
#> 468  -1.610143e+00      0
#> 469   6.521875e-01      0
#> 470   5.777866e-01      0
#> 471   3.384748e-01      0
#> 472  -6.489729e-01      0
#> 473   8.106252e-01      0
#> 474   1.439127e+00      0
#> 475   1.394430e+00      0
#> 476   1.706262e-02      0
#> 477  -1.051801e+00      0
#> 478  -1.829121e+00      0
#> 479  -8.502049e-02      0
#> 480   8.935684e-01      0
#> 481  -2.417850e+00      0
#> 482  -1.248170e+00      0
#> 483  -2.478727e+00      0
#> 484  -4.492382e-01      0
#> 485   4.664416e-01      0
#> 486   5.907391e-01      0
#> 487  -3.315229e-01      0
#> 488  -3.421311e-01      0
#> 489  -1.999913e+00      0
#> 490   3.331105e-02      0
#> 491   1.637051e+00      0
#> 492  -3.339950e-01      0
#> 493   3.327196e-01      0
#> 494  -3.527185e-01      0
#> 495   6.175127e-01      0
#> 496  -7.266549e-01      0
#> 497  -1.052076e-01      0
#> 498  -1.540873e-01      0
#> 499  -2.625772e+00      0
#> 500  -1.603969e+00      0
#> 501   2.034551e-01      0
#> 502   1.693161e+00      0
#> 503  -4.911509e-01      0
#> 504  -1.160472e+00      0
#> 505  -3.785720e-01      0
#> 506  -2.651868e-01      0
#> 507  -6.842930e-01      0
#> 508  -8.901762e-01      0
#> 509   3.947038e-01      0
#> 510  -1.249328e-01      0
#> 511  -2.006313e+00      0
#> 512   1.391965e-01      0
#> 513   1.319204e-01      0
#> 514   7.931299e-01      0
#> 515   5.232897e-01      0
#> 516  -5.205792e-01      0
#> 517   8.084117e-01      0
#> 518  -6.920312e-01      0
#> 519   2.410760e-01      0
#> 520  -3.009322e-01      0
#> 521  -7.583300e-02      0
#> 522  -5.206607e-01      0
#> 523   4.692239e-01      0
#> 524   5.566426e-01      0
#> 525   2.034113e-01      0
#> 526   2.089051e-01      0
#> 527  -1.136683e+00      0
#> 528  -5.215059e-02      0
#> 529   1.254283e+00      0
#> 530   4.455289e-01      0
#> 531  -2.598449e+00      0
#> 532  -6.621783e-01      0
#> 533  -1.136855e+00      0
#> 534   4.690251e-01      0
#> 535   1.574104e+00      0
#> 536  -1.682315e+00      0
#> 537  -2.686763e-01      0
#> 538  -1.133048e+00      0
#> 539   1.442643e+00      0
#> 540  -4.418639e-01      0
#> 541   7.102057e-01      0
#> 542   5.358389e-01      0
#> 543  -7.244057e-01      0
#> 544  -1.961610e+00      0
#> 545   1.307770e-02      0
#> 546   8.189400e-01      0
#> 547  -4.914435e-01      0
#> 548  -7.123861e-01      0
#> 549   1.067280e-01      0
#> 550   2.989697e-01      0
#> 551  -2.258798e-01      0
#> 552   6.312001e-01      0
#> 553   5.765923e-01      0
#> 554   7.819893e-01      0
#> 555  -1.336214e+00      0
#> 556   5.115758e-01      0
#> 557  -2.714514e-01      0
#> 558   4.727429e-01      0
#> 559  -1.355850e+00      0
#> 560  -1.653602e-01      0
#> 561   1.951398e-01      0
#> 562  -1.043076e+00      0
#> 563   7.658564e-01      0
#> 564   1.529984e+00      0
#> 565   1.804914e+00      0
#> 566  -1.061261e+00      0
#> 567   1.840266e+00      0
#> 568   6.790461e-01      0
#> 569  -6.324617e-01      0
#> 570  -1.204208e-01      0
#> 571  -1.095184e+00      0
#> 572   3.754582e-01      0
#> 573  -1.137449e+00      0
#> 574  -1.230446e+00      0
#> 575  -9.780012e-01      0
#> 576   7.977485e-01      0
#> 577  -1.934790e+00      0
#> 578   1.056730e+00      0
#> 579  -6.487129e-01      0
#> 580   6.306221e-01      0
#> 581  -4.428954e-01      0
#> 582   2.957168e-01      0
#> 583  -7.785157e-01      0
#> 584  -2.719413e-01      0
#> 585  -1.184210e+00      0
#> 586   1.384026e+00      0
#> 587  -1.153920e+00      0
#> 588  -4.426291e-02      0
#> 589  -6.158972e-01      0
#> 590   1.211936e+00      0
#> 591  -1.858486e-01      0
#> 592  -5.045557e-01      0
#> 593  -2.680119e+00      0
#> 594   8.928994e-01      0
#> 595   8.090818e-01      0
#> 596   2.869132e-01      0
#> 597   1.389995e+00      0
#> 598  -4.437511e-01      0
#> 599   1.800172e-01      0
#> 600   5.664436e-01      0
#> 601   2.276775e+00      0
#> 602   1.927109e-01      0
#> 603  -9.481445e-01      0
#> 604  -5.910594e-01      0
#> 605   4.168273e-01      0
#> 606  -1.139107e+00      0
#> 607  -7.743128e-01      0
#> 608   9.564624e-01      0
#> 609   5.374088e-01      0
#> 610  -1.334654e-01      0
#> 611   3.849492e-02      0
#> 612  -2.617274e-01      0
#> 613   6.009401e-01      0
#> 614  -1.052105e+00      0
#> 615   5.387947e-01      0
#> 616  -5.331395e-02      0
#> 617   5.951832e-01      0
#> 618   8.392296e-01      0
#> 619   7.453746e-01      0
#> 620  -1.020311e+00      0
#> 621   1.519796e-01      0
#> 622   2.791489e-01      0
#> 623  -1.194462e-01      0
#> 624   4.735107e-01      0
#> 625   2.521533e-01      0
#> 626  -2.121503e-01      0
#> 627   6.679904e-01      0
#> 628  -2.505546e+00      0
#> 629  -9.468905e-01      0
#> 630   5.264339e-01      0
#> 631   9.801981e-01      0
#> 632   1.494616e+00      0
#> 633  -1.190484e+00      0
#> 634   1.587645e+00      0
#> 635  -1.132795e+00      0
#> 636  -8.067431e-02      0
#> 637   5.886991e-01      0
#> 638   1.435653e+00      0
#> 639   4.115368e-01      0
#> 640   1.159521e-01      0
#> 641   3.364955e-01      0
#> 642   1.463053e+00      0
#> 643   2.551739e-01      0
#> 644  -8.896695e-01      0
#> 645  -1.179463e-01      0
#> 646   2.327719e+00      0
#> 647  -2.222753e-01      0
#> 648  -6.689496e-01      0
#> 649  -1.233400e+00      0
#> 650  -7.090870e-01      0
#> 651  -8.937313e-01      0
#> 652  -4.428949e-01      0
#> 653   1.464122e+00      0
#> 654   5.344345e-01      0
#> 655  -3.797864e-01      0
#> 656  -2.450236e-01      0
#> 657  -1.511686e+00      0
#> 658   1.665990e-01      0
#> 659   7.298886e-01      0
#> 660  -8.037115e-02      0
#> 661   1.824216e+00      0
#> 662  -2.291380e-01      0
#> 663  -2.990052e-01      0
#> 664   3.517549e-01      0
#> 665   1.411880e+00      0
#> 666  -2.333228e-01      0
#> 667   6.943087e-01      0
#> 668  -3.034979e-01      0
#> 669  -3.379651e-01      0
#> 670  -5.611911e-01      0
#> 671  -3.514831e-01      0
#> 672  -8.510294e-01      0
#> 673   1.320945e+00      0
#> 674  -1.217621e+00      0
#> 675   6.222593e-01      0
#> 676   4.508779e-01      0
#> 677  -6.305175e-01      0
#> 678   7.411121e-01      0
#> 679   5.121803e-01      0
#> 680   5.838127e-01      0
#> 681  -4.909997e-01      0
#> 682  -3.430922e-01      0
#> 683   2.388773e-01      0
#> 684   1.175095e+00      0
#> 685  -1.582200e+00      0
#> 686  -5.767130e-01      0
#> 687  -8.135325e-01      0
#> 688  -8.485209e-01      0
#> 689  -1.292440e-01      0
#> 690   3.869948e-01      0
#> 691   7.961412e-01      0
#> 692  -1.147380e+00      0
#> 693  -1.555028e-01      0
#> 694  -8.979816e-02      0
#> 695  -5.469739e-01      0
#> 696  -1.908851e-01      0
#> 697   2.607773e-01      0
#> 698  -7.618962e-01      0
#> 699   4.165644e-01      0
#> 700  -1.176747e-01      0
#> 701   1.462807e+00      0
#> 702   6.095700e-01      0
#> 703   7.470271e-02      0
#> 704   1.544624e+00      0
#> 705   1.419176e+00      0
#> 706  -4.310790e-01      0
#> 707  -1.273588e+00      0
#> 708  -1.088697e+00      0
#> 709  -1.971844e+00      0
#> 710  -9.621076e-01      0
#> 711  -4.319447e-01      0
#> 712   9.145806e-01      0
#> 713  -5.953511e-01      0
#> 714   1.624938e+00      0
#> 715  -5.927705e-01      0
#> 716  -8.638192e-01      0
#> 717  -9.049528e-02      0
#> 718   7.477522e-01      0
#> 719  -8.834679e-01      0
#> 720  -9.518018e-02      0
#> 721  -9.560355e-01      0
#> 722   4.892834e-01      0
#> 723  -1.425513e+00      0
#> 724   4.180776e-01      0
#> 725  -4.438314e-02      0
#> 726  -5.869097e-01      0
#> 727   6.098921e-01      0
#> 728  -5.336300e-01      0
#> 729  -1.184885e+00      0
#> 730  -2.250454e-01      0
#> 731   7.636138e-01      0
#> 732  -2.297265e+00      0
#> 733  -1.392983e+00      0
#> 734  -5.339862e-01      0
#> 735  -9.781748e-02      0
#> 736   2.436682e-01      0
#> 737  -8.482233e-01      0
#> 738   1.958239e+00      0
#> 739   2.195876e-02      0
#> 740   1.907178e-01      0
#> 741   4.684856e-01      0
#> 742  -7.493779e-01      0
#> 743  -1.678087e-01      0
#> 744  -1.226024e-01      0
#> 745  -4.691509e-01      0
#> 746  -6.721220e-01      0
#> 747   2.057114e+00      0
#> 748  -3.301381e-01      0
#> 749  -2.025279e+00      0
#> 750  -3.371400e-01      0
#> 751   7.191117e-02      0
#> 752   7.193415e-01      0
#> 753   7.154893e-03      0
#> 754   1.104737e+00      0
#> 755  -2.378359e-01      0
#> 756  -8.006334e-01      0
#> 757  -9.134931e-01      0
#> 758   1.117017e+00      0
#> 759   3.905063e-02      0
#> 760   3.470274e-01      0
#> 761   4.878840e-01      0
#> 762  -2.599263e+00      0
#> 763   1.967171e+00      0
#> 764  -3.514688e-01      0
#> 765   7.156213e-01      0
#> 766  -6.513207e-01      0
#> 767  -7.226663e-01      0
#> 768  -5.421507e-01      0
#> 769   1.897447e+00      0
#> 770   4.287834e-02      0
#> 771   1.618901e+00      0
#> 772   1.711158e+00      0
#> 773  -4.367116e-01      0
#> 774  -7.264272e-01      0
#> 775  -2.541815e+00      0
#> 776   1.180609e+00      0
#> 777   1.113675e+00      0
#> 778   1.251649e+00      0
#> 779  -1.470356e-01      0
#> 780  -4.892743e-01      0
#> 781   6.949084e-01      0
#> 782  -1.632748e+00      0
#> 783  -1.059059e+00      0
#> 784  -1.625086e-01      0
#> 785   6.536102e-01      0
#> 786   1.645797e+00      0
#> 787   1.944769e+00      0
#> 788   3.740163e-01      0
#> 789   1.554558e+00      0
#> 790  -2.399249e+00      0
#> 791   1.655806e+00      0
#> 792   1.081557e+00      0
#> 793   6.451341e-01      0
#> 794  -4.392737e-01      0
#> 795  -2.113010e-01      0
#> 796  -1.217322e+00      0
#> 797   6.586371e-01      0
#> 798   1.839105e-01      0
#> 799   2.495117e-01      0
#> 800  -2.495133e-01      0
#> 801  -3.693392e-01      0
#> 802  -3.726596e-01      0
#> 803  -1.346004e+00      0
#> 804   1.564228e+00      0
#> 805  -4.937272e-01      0
#> 806  -1.327347e-01      0
#> 807   3.389519e-01      0
#> 808   1.346960e+00      0
#> 809   2.296877e-01      0
#> 810   2.429789e-01      0
#> 811   1.137018e+00      0
#> 812  -2.212508e+00      0
#> 813  -1.384556e+00      0
#> 814  -7.150581e-02      0
#> 815  -1.623058e+00      0
#> 816   1.535165e+00      0
#> 817   7.622657e-01      0
#> 818  -4.797660e-01      0
#> 819  -2.713863e+00      0
#> 820   1.333432e+00      0
#> 821   9.154194e-01      0
#> 822   5.812506e-01      0
#> 823  -3.638001e-02      0
#> 824  -1.023508e+00      0
#> 825   1.083194e+00      0
#> 826  -1.348955e+00      0
#> 827   1.523385e+00      0
#> 828  -6.211769e-01      0
#> 829  -7.320764e-01      0
#> 830  -6.912559e-01      0
#> 831  -7.435150e-01      0
#> 832   7.985708e-01      0
#> 833  -3.466074e-01      0
#> 834   3.117715e-01      0
#> 835  -8.439650e-02      0
#> 836   1.568177e+00      0
#> 837  -3.909250e-01      0
#> 838  -2.534784e-01      0
#> 839  -2.616753e-01      0
#> 840   3.064486e-01      0
#> 841   9.751834e-01      0
#> 842   1.249970e+00      0
#> 843  -2.925743e-01      0
#> 844  -7.142850e-01      0
#> 845  -1.103347e+00      0
#> 846   1.288020e+00      0
#> 847  -9.445948e-01      0
#> 848  -1.939251e-01      0
#> 849  -1.139033e+00      0
#> 850  -1.614198e-01      0
#> 851   7.300445e-01      0
#> 852  -1.217172e-01      0
#> 853  -5.997206e-02      0
#> 854   3.577320e-01      0
#> 855  -5.084479e-01      0
#> 856   9.345905e-01      0
#> 857   2.966025e-01      0
#> 858  -1.230312e+00      0
#> 859  -8.306018e-01      0
#> 860   3.133786e-02      0
#> 861  -6.472201e-01      0
#> 862   3.511900e-01      0
#> 863  -7.255017e-01      0
#> 864  -3.897342e-01      0
#> 865   2.894007e-01      0
#> 866   5.215836e-01      0
#> 867   6.556420e-01      0
#> 868  -1.246444e+00      0
#> 869  -1.953900e+00      0
#> 870   5.632004e-01      0
#> 871  -4.413759e-01      0
#> 872  -2.091804e+00      0
#> 873   3.231675e-01      0
#> 874   1.100448e+00      0
#> 875   4.309659e-01      0
#> 876  -6.942166e-01      0
#> 877  -2.454324e-01      0
#> 878  -4.735305e-03      0
#> 879  -7.694761e-01      0
#> 880  -1.750214e+00      0
#> 881  -2.253518e+00      0
#> 882   9.962343e-01      0
#> 883  -5.980386e-01      0
#> 884   3.417527e-01      0
#> 885   2.997199e-02      0
#> 886  -1.684327e+00      0
#> 887   5.688219e-01      0
#> 888   4.442551e-01      0
#> 889  -1.095703e+00      0
#> 890  -2.371509e-01      0
#> 891  -1.308660e+00      0
#> 892   7.889944e-03      0
#> 893  -9.383308e-01      0
#> 894   8.098539e-01      0
#> 895  -6.884085e-01      0
#> 896  -2.178472e+00      0
#> 897  -1.827118e+00      0
#> 898  -7.672593e-02      0
#> 899  -5.941953e-01      0
#> 900   2.995644e-01      0
#> 901   3.597834e-01      0
#> 902   8.359816e-01      0
#> 903   5.528625e-01      0
#> 904   7.230907e-01      0
#> 905  -2.294390e-01      0
#> 906  -6.491326e-01      0
#> 907  -6.856922e-01      0
#> 908   2.200254e+00      0
#> 909  -2.023972e+00      0
#> 910   2.799794e-01      0
#> 911  -2.636723e-01      0
#> 912   8.458229e-01      0
#> 913   1.352651e+00      0
#> 914  -3.415205e-01      0
#> 915   6.485077e-02      0
#> 916   1.915843e-01      0
#> 917  -6.930811e-01      0
#> 918   8.031962e-01      1
#> 919   3.051603e-01      1
#> 920   1.646910e+00      1
#> 921  -1.199185e+00      1
#> 922   4.026426e-01      1
#> 923  -1.319660e+00      1
#> 924  -3.389198e-01      1
#> 925   5.667445e-01      1
#> 926   1.485108e-02      1
#> 927   7.560452e-01      1
#> 928   1.722915e+00      1
#> 929   1.129296e+00      1
#> 930   1.592686e+00      1
#> 931  -1.395882e+00      1
#> 932   2.215446e+00      1
#> 933  -1.445637e+00      1
#> 934   1.818343e+00      1
#> 935   8.881000e-01      1
#> 936   1.841566e+00      1
#> 937   8.505929e-01      1
#> 938  -7.976924e-01      1
#> 939  -2.064577e-02      1
#> 940   2.265247e-01      1
#> 941   2.033801e+00      1
#> 942   2.875400e-02      1
#> 943   6.163358e-01      1
#> 944   1.397756e+00      1
#> 945  -4.692698e-02      1
#> 946  -9.313774e-02      1
#> 947   1.239722e+00      1
#> 948   4.602339e-01      1
#> 949   2.159143e+00      1
#> 950  -2.422762e-01      1
#> 951   5.121516e-01      1
#> 952   1.245388e+00      1
#> 953   1.432465e+00      1
#> 954   2.390391e+00      1
#> 955   8.110732e-01      1
#> 956   1.828044e+00      1
#> 957   1.847000e+00      1
#> 958   1.715392e-01      1
#> 959   1.370319e+00      1
#> 960   3.151700e-01      1
#> 961  -4.358652e-01      1
#> 962   5.447298e-01      1
#> 963   1.080189e+00      1
#> 964   1.613847e+00      1
#> 965   1.726582e+00      1
#> 966   1.542962e+00      1
#> 967   7.775886e-02      1
#> 968   1.031567e+00      1
#> 969   1.164985e+00      1
#> 970   3.030008e+00      1
#> 971  -1.389214e+00      1
#> 972   8.140684e-01      1
#> 973   2.603273e+00      1
#> 974   1.177231e+00      1
#> 975   2.265625e+00      1
#> 976   7.172198e-02      1
#> 977  -1.052467e-01      1
#> 978   8.514212e-01      1
#> 979   5.402114e-01      1
#> 980   8.978763e-01      1
#> 981   3.289566e-01      1
#> 982   9.474615e-01      1
#> 983   2.332099e+00      1
#> 984   3.817840e-01      1
#> 985   4.881137e-01      1
#> 986   2.379802e+00      1
#> 987  -1.164231e+00      1
#> 988   9.232235e-01      1
#> 989   1.479320e+00      1
#> 990  -3.980385e-01      1
#> 991   1.267214e+00      1
#> 992  -7.822803e-01      1
#> 993   1.643974e+00      1
#> 994   8.500908e-01      1
#> 995   5.636872e-02      1
#> 996   1.142350e+00      1
#> 997   6.903800e-01      1
#> 998   1.833439e-01      1
#> 999   1.741696e+00      1
#> 1000  1.448751e+00      1