Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -0.3873951351      0
#> 2    -0.5148229662      0
#> 3     0.5192850568      0
#> 4    -0.5427408357      0
#> 5     1.9158427148      0
#> 6     0.2650723260      0
#> 7    -1.9166043213      0
#> 8     0.4188776552      0
#> 9    -0.1580831775      0
#> 10   -2.3702858250      0
#> 11    0.0587144655      0
#> 12    1.2287132076      0
#> 13   -0.0843161065      0
#> 14   -0.4620386640      0
#> 15   -0.4628676482      0
#> 16    0.5701886299      0
#> 17    0.9708444203      0
#> 18   -0.3362940393      0
#> 19    0.3605770345      0
#> 20    0.3167798532      0
#> 21    0.3382725676      0
#> 22    0.2777012690      0
#> 23    0.1770193051      0
#> 24   -0.5027225086      0
#> 25    0.3426516735      0
#> 26    0.6095267949      0
#> 27   -0.2428917833      0
#> 28   -0.6385549175      0
#> 29    0.8249746870      0
#> 30    1.1895456180      0
#> 31    0.9251197800      0
#> 32    0.2469603322      0
#> 33    1.0621833082      0
#> 34    0.3894603750      0
#> 35   -0.9357435878      0
#> 36   -1.1285643479      0
#> 37   -0.8923898987      0
#> 38    1.0495823841      0
#> 39    0.9425777175      0
#> 40   -0.6044787812      0
#> 41    0.6375234089      0
#> 42    0.2637633479      0
#> 43   -0.0709565946      0
#> 44    0.7198837903      0
#> 45    1.1225866710      0
#> 46   -0.1216545354      0
#> 47    0.7363379036      0
#> 48   -1.5933305742      0
#> 49    0.2674623641      0
#> 50   -0.5019064021      0
#> 51    1.1968420108      0
#> 52   -0.0150552297      0
#> 53   -0.6655728224      0
#> 54    0.9323336879      0
#> 55    0.8159669959      0
#> 56    0.2850902398      0
#> 57   -0.5965328034      0
#> 58    0.9581177669      0
#> 59    1.0067765121      0
#> 60    0.4479870767      0
#> 61    0.6686812809      0
#> 62    0.8707492613      0
#> 63    0.4194960439      0
#> 64    0.7052946594      0
#> 65   -0.5691532645      0
#> 66    1.2122890995      0
#> 67   -1.4532272690      0
#> 68    0.7391931094      0
#> 69    0.1530047931      0
#> 70    1.0914455683      0
#> 71   -0.4045900605      0
#> 72    0.7773258899      0
#> 73   -0.1439493097      0
#> 74   -1.1697063137      0
#> 75   -1.4661215053      0
#> 76   -0.4087651287      0
#> 77   -1.7292039316      0
#> 78   -0.1385477690      0
#> 79    0.3865775649      0
#> 80    0.7211115405      0
#> 81    0.1891414082      0
#> 82   -1.2374950170      0
#> 83    0.2585593234      0
#> 84    1.6447428476      0
#> 85    0.4266774097      0
#> 86    0.4298201293      0
#> 87    0.0247851345      0
#> 88   -0.3583569624      0
#> 89    2.0465255519      0
#> 90    0.7514689705      0
#> 91    0.8388788380      0
#> 92    0.4486104205      0
#> 93    0.0720569608      0
#> 94    0.3320208049      0
#> 95   -1.0305974517      0
#> 96    0.0323045061      0
#> 97   -1.7990286776      0
#> 98   -1.2949481350      0
#> 99   -1.6562900946      0
#> 100   1.4977802377      0
#> 101   0.0787597599      0
#> 102  -1.6863701694      0
#> 103   0.2733628381      0
#> 104  -0.4756473034      0
#> 105   0.9319880693      0
#> 106   1.4279726719      0
#> 107  -1.1045698424      0
#> 108  -1.4799048764      0
#> 109   0.7674789413      0
#> 110  -1.1848738414      0
#> 111   0.6108221840      0
#> 112   0.4093654836      0
#> 113   0.1825938058      0
#> 114  -0.9063436129      0
#> 115   0.3554272693      0
#> 116   1.5631994874      0
#> 117  -0.1935705062      0
#> 118   1.0279483700      0
#> 119  -0.9468176585      0
#> 120  -1.0603772126      0
#> 121  -0.8497970230      0
#> 122  -0.9136621786      0
#> 123  -0.7169159889      0
#> 124  -0.3158465059      0
#> 125  -1.0235440941      0
#> 126   1.9701698764      0
#> 127  -1.3669065848      0
#> 128  -0.6049911998      0
#> 129  -1.1025258958      0
#> 130   0.7483412367      0
#> 131  -1.1114447558      0
#> 132   0.5087284401      0
#> 133   0.0693398846      0
#> 134  -1.0950749237      0
#> 135  -0.8632833183      0
#> 136  -0.3309313854      0
#> 137  -0.5569972123      0
#> 138  -0.9618421760      0
#> 139   0.4932489988      0
#> 140  -0.1538248699      0
#> 141  -0.9799957853      0
#> 142  -0.2737031158      0
#> 143   0.4351692130      0
#> 144   0.5320103895      0
#> 145  -0.7221218505      0
#> 146   1.8970529268      0
#> 147  -0.7542303679      0
#> 148   0.8032359722      0
#> 149   0.1045041211      0
#> 150   2.5847936920      0
#> 151   1.3657768663      0
#> 152   0.1302760689      0
#> 153  -0.4502704614      0
#> 154  -2.2454614788      0
#> 155  -0.9497361255      0
#> 156  -0.8340137428      0
#> 157   0.3446617589      0
#> 158  -0.0789883564      0
#> 159  -0.7978244755      0
#> 160   0.3590812418      0
#> 161   1.0160744786      0
#> 162   0.5706667668      0
#> 163  -0.9049546762      0
#> 164  -0.2007152446      0
#> 165   0.2847599159      0
#> 166  -0.4605285314      0
#> 167   1.1321417353      0
#> 168  -0.2280237548      0
#> 169   0.3157204041      0
#> 170   1.6569830716      0
#> 171   0.4397941181      0
#> 172   1.0591852856      0
#> 173   0.3235402988      0
#> 174  -0.3141652046      0
#> 175   1.0734061792      0
#> 176  -0.6610399243      0
#> 177  -0.2782010472      0
#> 178   0.3018639006      0
#> 179  -0.4385286135      0
#> 180   0.7358234527      0
#> 181   0.7846932994      0
#> 182   0.0727009395      0
#> 183  -0.3790856112      0
#> 184   0.8825752110      0
#> 185   1.4890444526      0
#> 186  -0.9668310914      0
#> 187  -1.7803000634      0
#> 188  -0.9662524286      0
#> 189  -0.0337790742      0
#> 190  -0.7416019758      0
#> 191   1.0856062294      0
#> 192  -0.3616824631      0
#> 193   0.4744139444      0
#> 194   0.7415358579      0
#> 195   0.8587905130      0
#> 196   1.8109344190      0
#> 197   0.1005836321      0
#> 198  -0.7258967384      0
#> 199   1.7148217015      0
#> 200  -0.6064456788      0
#> 201  -1.1335716250      0
#> 202  -0.4354209708      0
#> 203   0.8686553652      0
#> 204  -1.9409952997      0
#> 205   0.1233804134      0
#> 206   0.1749430688      0
#> 207   0.6158080263      0
#> 208  -0.5883327761      0
#> 209  -0.1908317331      0
#> 210  -1.5072462316      0
#> 211  -1.4105865988      0
#> 212  -0.3078738174      0
#> 213   0.4364230295      0
#> 214  -0.0289903283      0
#> 215  -1.8776660963      0
#> 216   0.2776505429      0
#> 217   0.8929888408      0
#> 218   0.0261854256      0
#> 219   0.1184469088      0
#> 220  -0.9633939849      0
#> 221  -0.2240025151      0
#> 222  -0.1292301494      0
#> 223  -0.2049509830      0
#> 224   0.2009701167      0
#> 225  -2.0841948635      0
#> 226   2.0640604467      0
#> 227   0.7041050397      0
#> 228  -0.2000743843      0
#> 229  -0.9895309502      0
#> 230  -0.4201928622      0
#> 231  -0.4054815727      0
#> 232  -0.5512253152      0
#> 233   0.1558547394      0
#> 234   0.0730398891      0
#> 235   0.8569338388      0
#> 236   0.4913608406      0
#> 237  -1.0493070088      0
#> 238   0.3121066662      0
#> 239   0.7113119568      0
#> 240   0.3486631705      0
#> 241   1.0398763257      0
#> 242   2.0031925343      0
#> 243   0.2159061730      0
#> 244   1.3139451480      0
#> 245  -0.6489645835      0
#> 246  -0.6366355122      0
#> 247   0.4048215903      0
#> 248  -0.7884340243      0
#> 249   2.0532444309      0
#> 250   0.6630509633      0
#> 251  -0.2077659386      0
#> 252  -0.4500050445      0
#> 253   0.1962869929      0
#> 254   0.1979771352      0
#> 255  -1.6372543376      0
#> 256   0.8772055558      0
#> 257  -0.2868191434      0
#> 258  -1.7928262890      0
#> 259   0.1557624945      0
#> 260  -0.5579395924      0
#> 261   1.6462605061      0
#> 262   0.3199642894      0
#> 263  -0.8155405904      0
#> 264  -0.3026532887      0
#> 265  -0.0025557945      0
#> 266  -0.3911825527      0
#> 267   0.8400720943      0
#> 268  -0.3364418816      0
#> 269  -0.8875001381      0
#> 270   0.0616157778      0
#> 271  -0.3349436684      0
#> 272   2.5741614458      0
#> 273   2.5625679829      0
#> 274  -0.5081044684      0
#> 275   1.2839898211      0
#> 276   0.5349161400      0
#> 277  -1.8377316035      0
#> 278   0.2806293781      0
#> 279   0.7181259457      0
#> 280   0.3185582126      0
#> 281   1.6146941572      0
#> 282  -0.6908140382      0
#> 283   0.8172600464      0
#> 284  -1.5754757452      0
#> 285  -0.0187877989      0
#> 286  -1.2154328472      0
#> 287   0.7718946378      0
#> 288   0.5604359139      0
#> 289   2.1776776517      0
#> 290  -1.3101517407      0
#> 291  -2.2419076483      0
#> 292   0.7465124537      0
#> 293   1.8088749068      0
#> 294  -0.3706181371      0
#> 295  -0.2214185817      0
#> 296   0.8884571286      0
#> 297   0.8552330720      0
#> 298   1.2217823362      0
#> 299   0.2636636141      0
#> 300   0.0185676134      0
#> 301  -1.8109948264      0
#> 302   1.4753873099      0
#> 303   0.0552308782      0
#> 304   0.8627578446      0
#> 305  -0.5627330269      0
#> 306   1.0504074719      0
#> 307   0.4350617400      0
#> 308   1.8485832449      0
#> 309  -0.7438641343      0
#> 310   1.0193344581      0
#> 311   0.2165939631      0
#> 312  -0.3301465233      0
#> 313  -0.5772517520      0
#> 314  -0.1877453041      0
#> 315  -1.6046348672      0
#> 316   0.7373293340      0
#> 317   0.6089653382      0
#> 318   1.6249477193      0
#> 319  -0.2824374198      0
#> 320   0.7793644346      0
#> 321  -0.0742940526      0
#> 322   1.6077877993      0
#> 323  -0.4771377820      0
#> 324   0.5072472586      0
#> 325  -2.2021800637      0
#> 326   0.3216762114      0
#> 327   0.5767910989      0
#> 328   0.5229749009      0
#> 329  -0.3136223687      0
#> 330  -0.7201576758      0
#> 331   0.5556708602      0
#> 332  -0.6343716410      0
#> 333  -0.0049194950      0
#> 334  -0.1732669701      0
#> 335  -0.0877717682      0
#> 336  -0.1487882993      0
#> 337   0.3233326964      0
#> 338   1.1607253436      0
#> 339  -0.7466272582      0
#> 340  -0.4653853961      0
#> 341   0.5550975672      0
#> 342  -1.3070671356      0
#> 343   0.3421886261      0
#> 344   0.9770809183      0
#> 345  -1.0966849574      0
#> 346  -1.5400864899      0
#> 347  -0.6223202359      0
#> 348   0.7517885278      0
#> 349  -0.3574427530      0
#> 350   1.4995802069      0
#> 351  -0.0592685401      0
#> 352   0.1316942501      0
#> 353  -1.6102224603      0
#> 354   2.1940067629      0
#> 355   0.1120273266      0
#> 356  -0.1840516256      0
#> 357  -0.5949809408      0
#> 358   0.7321287537      0
#> 359  -1.4396035416      0
#> 360   0.5827900480      0
#> 361   0.7821866399      0
#> 362   0.1179637239      0
#> 363  -0.6778090201      0
#> 364   0.1690938492      0
#> 365   1.2826248844      0
#> 366   0.4120980395      0
#> 367   0.5253365725      0
#> 368  -0.2444126151      0
#> 369   1.4308297009      0
#> 370  -0.2712636782      0
#> 371   0.1230259575      0
#> 372   0.5559096443      0
#> 373   0.3023566395      0
#> 374   1.3769362189      0
#> 375  -0.2730353487      0
#> 376   0.3106254834      0
#> 377  -0.2155749061      0
#> 378  -0.5821417071      0
#> 379  -0.4582476460      0
#> 380   0.7987585311      0
#> 381   0.1599672870      0
#> 382   1.3748735382      0
#> 383   0.4436725440      0
#> 384   2.4247951956      0
#> 385   2.7163384524      0
#> 386  -1.3574540018      0
#> 387  -0.4359926456      0
#> 388  -0.7473774301      0
#> 389  -0.6177044849      0
#> 390  -0.7949554098      0
#> 391   0.5887918049      0
#> 392  -1.4466289496      0
#> 393  -0.8756829385      0
#> 394  -1.1809951119      0
#> 395   0.9385950267      0
#> 396   0.2465499064      0
#> 397   0.0281090126      0
#> 398   0.2046579034      0
#> 399   1.0819949394      0
#> 400   0.3789857090      0
#> 401  -1.2760413854      0
#> 402   0.3788964435      0
#> 403  -0.2020362017      0
#> 404  -0.6618713485      0
#> 405  -0.4137930079      0
#> 406  -0.4004676080      0
#> 407   0.4368074685      0
#> 408  -1.7240494963      0
#> 409   0.4334413139      0
#> 410  -1.4199582996      0
#> 411  -0.5538172119      0
#> 412   2.0854286128      0
#> 413   0.4350114940      0
#> 414   0.4029877931      0
#> 415   1.3507158033      0
#> 416   0.3205203507      0
#> 417   0.0130812720      0
#> 418   0.5898490845      0
#> 419   1.5302635867      0
#> 420  -2.0817477583      0
#> 421   1.2395683239      0
#> 422   1.7455189910      0
#> 423  -0.3578945816      0
#> 424   0.2465479619      0
#> 425  -0.2522371781      0
#> 426   0.3009090851      0
#> 427   0.9126780608      0
#> 428  -1.5964328957      0
#> 429   0.8938010973      0
#> 430   1.4644850216      0
#> 431   0.1757360694      0
#> 432  -0.0769943843      0
#> 433   0.1860473006      0
#> 434  -0.3784558962      0
#> 435  -0.5703273854      0
#> 436  -0.7082617199      0
#> 437  -0.3711107494      0
#> 438  -0.3880070640      0
#> 439  -0.3788921096      0
#> 440   0.5238475577      0
#> 441   0.1554442861      0
#> 442  -1.0085154356      0
#> 443   0.0459729294      0
#> 444   0.2849195775      0
#> 445   0.8565542478      0
#> 446  -0.7277471739      0
#> 447  -0.8566021651      0
#> 448   0.2662646886      0
#> 449   1.3534873457      0
#> 450  -1.7575618760      0
#> 451   0.5594441606      0
#> 452   0.6150545734      0
#> 453  -0.5176017598      0
#> 454  -0.8691561549      0
#> 455  -1.6826058006      0
#> 456  -1.2746235500      0
#> 457  -0.0062593991      0
#> 458   0.3140690950      0
#> 459   0.7100449836      0
#> 460  -1.7914038696      0
#> 461   2.5595631753      0
#> 462  -0.4655237737      0
#> 463  -1.0370766882      0
#> 464  -1.0149430244      0
#> 465  -0.6661254393      0
#> 466   0.9905166312      0
#> 467  -0.2723799040      0
#> 468  -0.2592568201      0
#> 469  -0.8836769536      0
#> 470  -0.5790622747      0
#> 471   2.2022587721      0
#> 472   0.0501023476      0
#> 473   1.6006360138      0
#> 474   0.4243912571      0
#> 475   0.6291708750      0
#> 476   1.6861630821      0
#> 477   2.6182029270      0
#> 478  -0.7469709762      0
#> 479   0.4545926220      0
#> 480  -0.4337825667      0
#> 481  -0.1874407525      0
#> 482  -1.2420693538      0
#> 483  -0.2032310132      0
#> 484   0.3310202292      0
#> 485  -0.2378895826      0
#> 486   0.7673607983      0
#> 487   1.6496386920      0
#> 488   1.5469804448      0
#> 489  -0.9399571252      0
#> 490   0.0003960717      0
#> 491   0.2783855800      0
#> 492   0.8837205592      0
#> 493   1.1884515928      0
#> 494   2.7774211796      0
#> 495   0.6248327249      0
#> 496   1.8263645516      0
#> 497   1.6741094236      0
#> 498  -0.4948517457      0
#> 499  -0.6977705094      0
#> 500   1.1680373704      0
#> 501   0.8427739453      0
#> 502   0.5533058382      0
#> 503  -1.1311275169      0
#> 504   0.8534236472      0
#> 505   0.2987411404      0
#> 506  -0.1861373430      0
#> 507   0.1178203512      0
#> 508   1.4060369170      0
#> 509   0.5176822817      0
#> 510  -1.0046475631      0
#> 511  -0.0649474092      0
#> 512  -0.7221446485      0
#> 513  -0.1603849000      0
#> 514   0.7400657252      0
#> 515  -0.0775822190      0
#> 516  -1.1930727092      0
#> 517   0.5409477643      0
#> 518  -0.6938434114      0
#> 519   1.9542234625      0
#> 520  -0.1707667439      0
#> 521  -0.0090937030      0
#> 522  -1.7956122000      0
#> 523  -1.6706688314      0
#> 524  -1.2273376575      0
#> 525   0.3383849621      0
#> 526   1.4070760608      0
#> 527   1.2291155486      0
#> 528  -0.0066737748      0
#> 529   2.0019575778      0
#> 530   0.1994152028      0
#> 531   0.8507701967      0
#> 532   1.3894157724      0
#> 533  -1.2269357886      0
#> 534  -1.1415914764      0
#> 535   0.4835841769      0
#> 536  -1.1067908188      0
#> 537  -0.8725263424      0
#> 538  -0.8963776356      0
#> 539   0.3033502676      0
#> 540   1.2435383605      0
#> 541   0.7690084390      0
#> 542  -0.6850386609      0
#> 543  -1.2533892666      0
#> 544  -0.6707375502      0
#> 545   0.4652578245      0
#> 546   0.3299339316      0
#> 547  -0.2795211834      0
#> 548  -0.0024445867      0
#> 549   0.7764512459      0
#> 550  -3.0170570647      0
#> 551   1.2404877942      0
#> 552  -1.4978650237      0
#> 553   1.0565631736      0
#> 554  -1.4387140044      0
#> 555  -0.6776417308      0
#> 556  -0.1775221150      0
#> 557  -1.2945047030      0
#> 558  -2.0558267625      0
#> 559  -0.9267280811      0
#> 560   1.5037521292      0
#> 561   0.3885446271      0
#> 562  -0.5206910593      0
#> 563   0.1484852860      0
#> 564  -1.2830777530      0
#> 565   1.4271609221      0
#> 566   0.9896725935      0
#> 567  -0.0077733024      0
#> 568  -0.5622039232      0
#> 569   0.9902563216      0
#> 570  -0.9402353609      0
#> 571   0.9122612319      0
#> 572  -0.5590559716      0
#> 573  -0.3094498258      0
#> 574   0.6516197232      0
#> 575   0.4672413834      0
#> 576   0.1158195147      0
#> 577  -0.1653106301      0
#> 578   0.0385854951      0
#> 579  -0.8281870620      0
#> 580  -0.9421759784      0
#> 581   0.3106940640      0
#> 582   0.3260417084      0
#> 583   0.3277458015      0
#> 584  -0.5611815642      0
#> 585   0.3582700519      0
#> 586   0.7750149355      0
#> 587   0.5167533536      0
#> 588   0.5379377814      0
#> 589   0.1346021663      0
#> 590   2.0841224191      0
#> 591  -0.6703526183      0
#> 592   1.5073619300      0
#> 593  -1.1891091454      0
#> 594  -2.2719433628      0
#> 595   0.2215526908      0
#> 596  -0.5291875023      0
#> 597   0.1886499480      0
#> 598  -2.0406511030      0
#> 599  -0.8515623363      0
#> 600  -0.0152886152      0
#> 601   0.8694741855      0
#> 602  -1.4402550328      0
#> 603   2.2655738092      0
#> 604   1.4613887043      0
#> 605  -0.8190674769      0
#> 606   1.5142417821      0
#> 607  -0.1959252043      0
#> 608   0.8405734847      0
#> 609   0.0730766112      0
#> 610   0.3064399884      0
#> 611  -0.4971786588      0
#> 612  -0.1071432531      0
#> 613   1.4710211270      0
#> 614   0.3687193053      0
#> 615  -0.1350488151      0
#> 616   0.6263773756      0
#> 617  -2.0182879339      0
#> 618   0.4144778083      0
#> 619  -0.1271793904      0
#> 620   0.4802722131      0
#> 621   1.2741690469      0
#> 622  -0.3296151723      0
#> 623   0.2739790396      0
#> 624   0.0239002521      0
#> 625   0.0333324123      0
#> 626   0.6155103083      0
#> 627   1.1967042045      0
#> 628   2.2717469046      0
#> 629   2.2286325600      0
#> 630   0.1428209441      0
#> 631   0.2095039156      0
#> 632   0.0672991017      0
#> 633  -2.4731020267      0
#> 634   0.7856319688      0
#> 635  -1.4996384718      0
#> 636   1.2145885035      0
#> 637  -0.5453616841      0
#> 638   1.0179493474      0
#> 639  -0.5015060274      0
#> 640  -0.2980215731      0
#> 641  -0.9887829652      0
#> 642   1.3275953369      0
#> 643  -0.5989141407      0
#> 644   0.6524617111      0
#> 645   0.4103452341      0
#> 646  -0.0986324223      0
#> 647  -2.0323861768      0
#> 648  -0.5921615102      0
#> 649  -0.7888028768      0
#> 650   0.2268361247      0
#> 651   0.3439859973      0
#> 652  -0.2200610025      0
#> 653  -1.1289514149      0
#> 654  -0.0733178985      0
#> 655  -0.8206310218      0
#> 656  -0.3260778608      0
#> 657   1.3671630574      0
#> 658   0.8584313915      0
#> 659   0.3029601379      0
#> 660   0.0292714476      0
#> 661  -0.6145920935      0
#> 662  -0.1232565869      0
#> 663  -1.1136459971      0
#> 664  -1.6703105101      0
#> 665   1.1284729722      0
#> 666  -1.1764386698      0
#> 667  -0.0467549765      0
#> 668  -1.5991222628      0
#> 669  -0.0575204468      0
#> 670   0.6190861090      0
#> 671   0.4436891168      0
#> 672  -0.0990143548      0
#> 673  -0.4114734372      0
#> 674  -1.4125166604      0
#> 675   0.3984298447      0
#> 676   1.6765817452      0
#> 677   0.4920281091      0
#> 678   0.9823544235      0
#> 679   1.1513251221      0
#> 680  -1.4557285517      0
#> 681   0.2541721601      0
#> 682  -0.8213721733      0
#> 683   0.6079890031      0
#> 684   0.9891746416      0
#> 685   0.2765643012      0
#> 686   0.8873882560      0
#> 687   0.0529606470      0
#> 688   0.6472493709      0
#> 689  -1.8389723455      0
#> 690   0.8084904589      0
#> 691  -0.5617457483      0
#> 692   0.8873966438      0
#> 693   0.8632807820      0
#> 694  -0.6195861214      0
#> 695  -0.2692849549      0
#> 696   1.9031225298      0
#> 697   0.2019004370      0
#> 698   0.4734897111      0
#> 699   2.3055380829      0
#> 700   0.4808094224      0
#> 701   0.1427172300      0
#> 702   1.8387236364      0
#> 703  -1.1797822751      0
#> 704  -0.6646030334      0
#> 705   0.0263866419      0
#> 706  -1.5834389519      0
#> 707  -0.5332869351      0
#> 708   1.2260846649      0
#> 709   0.4086528543      0
#> 710   0.2307882325      0
#> 711  -0.4331389133      0
#> 712  -1.3300413682      0
#> 713  -0.4482484729      0
#> 714   1.5965692585      0
#> 715  -0.6688440662      0
#> 716  -0.7409672867      0
#> 717   0.7163697275      0
#> 718   0.5413895006      0
#> 719   0.0585307840      0
#> 720  -1.4890075823      0
#> 721   0.3087778973      0
#> 722  -0.9828937851      0
#> 723  -0.3824669766      0
#> 724  -0.5289124176      0
#> 725   0.1335201985      0
#> 726  -0.2988045553      0
#> 727   0.4137148098      0
#> 728  -0.1302080605      0
#> 729   0.8206112044      0
#> 730  -0.7404911024      0
#> 731  -0.2006047342      0
#> 732   0.0784570616      0
#> 733   0.8159301618      0
#> 734  -0.8473430521      0
#> 735  -0.9037094860      0
#> 736  -0.8342135340      0
#> 737   0.5644208203      0
#> 738   1.0377802179      0
#> 739   0.5056792414      0
#> 740   0.5599776855      0
#> 741   1.3765029255      0
#> 742   1.3328347519      0
#> 743  -1.5179551425      0
#> 744  -0.5431109763      0
#> 745  -0.5890152143      0
#> 746  -1.2252855637      0
#> 747   0.0593426763      0
#> 748   0.1134442526      0
#> 749   1.1421549887      0
#> 750  -0.0216402021      0
#> 751  -0.0966185897      0
#> 752   0.7319667972      0
#> 753  -1.2389105601      0
#> 754   0.4528057334      0
#> 755  -1.2424037216      0
#> 756   0.2492725475      0
#> 757  -0.2226945436      0
#> 758   0.5909948757      0
#> 759  -0.9255675039      0
#> 760   1.1092464227      0
#> 761   0.0808312413      0
#> 762  -0.6764602313      0
#> 763  -1.7342780131      0
#> 764  -0.4657369734      0
#> 765  -0.1106176713      0
#> 766   0.1518318735      0
#> 767   0.5207076025      0
#> 768   0.7198967335      0
#> 769   1.1379031371      0
#> 770   0.3905392831      0
#> 771  -1.5331922291      0
#> 772  -0.8826313362      0
#> 773  -0.0901715998      0
#> 774  -1.7294240526      0
#> 775   0.5913971501      0
#> 776  -0.4838469076      0
#> 777   2.0274275420      0
#> 778   0.4769962511      0
#> 779   0.6128954433      0
#> 780   0.5436799001      0
#> 781   0.2284626680      0
#> 782   1.5149894011      0
#> 783   1.0292717505      0
#> 784   1.2108082917      0
#> 785   0.1483757448      0
#> 786  -0.5259256943      0
#> 787  -1.1753838733      0
#> 788  -0.4227077842      0
#> 789   0.7368239376      0
#> 790   0.1887065288      0
#> 791   0.2110944549      0
#> 792  -0.8163169979      0
#> 793  -1.0748598390      0
#> 794   0.7163458173      0
#> 795   1.0445084266      0
#> 796  -0.4655750439      0
#> 797   0.3855131030      0
#> 798  -0.3913518387      0
#> 799   0.1362629526      0
#> 800   0.5732315359      0
#> 801   1.4964848706      0
#> 802   0.1942209156      0
#> 803   0.6943125544      0
#> 804   1.8372703742      0
#> 805   1.0353062179      0
#> 806   0.7294325991      0
#> 807  -0.5867568654      0
#> 808   0.5743885452      0
#> 809  -0.2850301185      0
#> 810  -2.3496405230      0
#> 811  -0.7827799169      0
#> 812   0.9548762914      0
#> 813  -0.0032594973      0
#> 814  -0.4113134441      0
#> 815   0.5892106699      0
#> 816  -0.7759778038      0
#> 817  -0.4123558687      0
#> 818  -0.0948672759      0
#> 819   0.7750160528      0
#> 820  -0.8889086121      0
#> 821  -1.3200648153      0
#> 822  -1.2675445162      0
#> 823   0.3090470383      0
#> 824  -0.6522238038      0
#> 825   1.6388673055      0
#> 826  -1.0958223090      0
#> 827  -0.3578922121      0
#> 828  -0.2638697648      0
#> 829   0.9079158428      0
#> 830   0.5305863341      0
#> 831   0.0001337948      0
#> 832  -0.7411112804      0
#> 833  -2.3157343467      0
#> 834   1.0116562604      0
#> 835   0.0935949846      0
#> 836   0.1763516096      0
#> 837  -0.7039234456      0
#> 838  -0.8218501607      0
#> 839  -0.6771502746      0
#> 840   1.2032784575      0
#> 841  -0.0427962157      0
#> 842  -0.2094185502      0
#> 843  -0.0931255760      0
#> 844   1.5645113883      0
#> 845  -0.9093864548      0
#> 846  -0.2500013314      0
#> 847   0.4014556940      0
#> 848  -0.6070104316      0
#> 849  -1.1260538481      0
#> 850   1.0107069398      0
#> 851  -1.2440570204      0
#> 852   0.8275134541      0
#> 853  -1.6031481197      0
#> 854  -1.4171774957      0
#> 855   0.8168827048      0
#> 856   1.1510540090      0
#> 857  -0.5472344911      0
#> 858  -0.5897418226      0
#> 859   0.1598421702      0
#> 860   1.2652090593      0
#> 861   0.8352497109      0
#> 862   0.6926749420      0
#> 863   0.4874218314      0
#> 864  -0.1588746388      0
#> 865   1.0587545790      0
#> 866  -0.0918627591      0
#> 867  -0.7244357725      0
#> 868   0.1777673167      0
#> 869  -0.1454126149      0
#> 870   0.6463010440      0
#> 871  -0.9713007064      0
#> 872   0.8540902641      0
#> 873  -0.7139743265      0
#> 874   0.3967451785      0
#> 875   0.0212431771      0
#> 876   0.8973322585      0
#> 877   0.6985542120      0
#> 878   0.1429130773      0
#> 879  -0.2149064064      0
#> 880  -1.9564001229      0
#> 881   1.5448744306      0
#> 882  -0.2369102360      0
#> 883  -1.9705329601      0
#> 884  -0.6901422784      0
#> 885   0.3726041521      0
#> 886  -0.8325515955      0
#> 887   1.6156437910      0
#> 888  -0.4216552576      0
#> 889  -0.6675670796      0
#> 890  -0.1821284621      0
#> 891   0.7645377112      0
#> 892  -1.2509585797      0
#> 893  -0.3794432197      0
#> 894   0.7924124415      0
#> 895   0.9079705092      0
#> 896  -0.4597456086      0
#> 897   1.2231554167      1
#> 898  -0.5034838657      1
#> 899   0.9290009566      1
#> 900   0.2438523598      1
#> 901  -1.0424756277      1
#> 902   0.3337725493      1
#> 903  -0.0322899725      1
#> 904   2.4404475565      1
#> 905   2.6979649656      1
#> 906   0.9494737439      1
#> 907   1.3648241475      1
#> 908   0.6201938880      1
#> 909   2.3892789604      1
#> 910  -0.7238194936      1
#> 911   0.5731986635      1
#> 912   0.1885723173      1
#> 913   0.6408444232      1
#> 914   1.4675428377      1
#> 915   0.8243760403      1
#> 916   1.5827238514      1
#> 917  -0.1156789160      1
#> 918   0.0807480150      1
#> 919  -0.9371033890      1
#> 920   3.0049525292      1
#> 921   0.6152808348      1
#> 922   1.7552860087      1
#> 923   2.5861151755      1
#> 924   1.0747317566      1
#> 925   0.2445207733      1
#> 926   1.9653564086      1
#> 927   1.5494050242      1
#> 928   2.5064739717      1
#> 929   0.4699714941      1
#> 930   1.7823370947      1
#> 931   1.0410586487      1
#> 932   0.3149509313      1
#> 933   1.6744053830      1
#> 934  -1.0612945376      1
#> 935  -0.6201403943      1
#> 936   2.0144626447      1
#> 937   3.1967929596      1
#> 938   0.2510437304      1
#> 939   1.5211575547      1
#> 940   1.6115062081      1
#> 941  -1.0766700448      1
#> 942   0.2827201146      1
#> 943   0.4011750810      1
#> 944   0.1553483394      1
#> 945  -0.5523753242      1
#> 946   1.9392604970      1
#> 947   1.0274863079      1
#> 948   1.2110665500      1
#> 949   2.8914794712      1
#> 950   1.8061073172      1
#> 951   0.0794158195      1
#> 952   1.0997966387      1
#> 953   0.4212337945      1
#> 954   1.3972183211      1
#> 955   1.9419666795      1
#> 956  -0.2617977829      1
#> 957   1.2777089359      1
#> 958  -0.0339196493      1
#> 959   1.8066743544      1
#> 960   0.7825468141      1
#> 961   0.8567765355      1
#> 962   0.1706005922      1
#> 963   1.4764863489      1
#> 964   0.3379858337      1
#> 965   0.9077290992      1
#> 966   0.6546894304      1
#> 967   0.3548858809      1
#> 968  -0.2770129564      1
#> 969  -0.4347163324      1
#> 970   1.6617471603      1
#> 971   1.0379172734      1
#> 972   1.0858590462      1
#> 973   0.5407606058      1
#> 974   2.5762208469      1
#> 975   1.4494772026      1
#> 976   1.4222060494      1
#> 977   1.6942847944      1
#> 978   0.2666113073      1
#> 979   1.6613976969      1
#> 980   0.5113556188      1
#> 981   1.3490890150      1
#> 982   0.7387534416      1
#> 983   0.7576129869      1
#> 984   1.2604440859      1
#> 985   1.7168698201      1
#> 986   1.7330359814      1
#> 987   0.0605222198      1
#> 988   0.6452049693      1
#> 989   0.4530465844      1
#> 990   2.6015051407      1
#> 991   0.2622080597      1
#> 992  -0.0749861032      1
#> 993  -0.7899544756      1
#> 994  -0.7054040753      1
#> 995  -0.4854750905      1
#> 996   1.3241474956      1
#> 997   1.1245196172      1
#> 998   0.5105217938      1
#> 999  -0.2973843260      1
#> 1000  1.3137245440      1