Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     1.9730815701      0
#> 2    -0.8976228690      0
#> 3     0.6798668029      0
#> 4    -0.2984012920      0
#> 5    -0.3880659212      0
#> 6    -0.8837415657      0
#> 7     1.1541737216      0
#> 8     0.9459107368      0
#> 9    -1.3460957083      0
#> 10    1.5480456175      0
#> 11    0.6814351134      0
#> 12   -0.6106768916      0
#> 13    0.3339046220      0
#> 14    0.1447210627      0
#> 15    0.3092282143      0
#> 16   -0.6334416963      0
#> 17   -0.6135472898      0
#> 18    1.0524023848      0
#> 19   -0.1972658229      0
#> 20    2.3673644913      0
#> 21    0.4774805716      0
#> 22   -0.2662299409      0
#> 23    0.5938985552      0
#> 24   -0.1092494895      0
#> 25   -0.0942173177      0
#> 26   -0.5292866193      0
#> 27   -0.5171928657      0
#> 28   -1.0276424288      0
#> 29    0.4191885749      0
#> 30    0.1825764413      0
#> 31   -0.2823657992      0
#> 32   -2.5232643519      0
#> 33   -0.0547199593      0
#> 34    0.1663623299      0
#> 35   -1.4843086590      0
#> 36   -0.3888021135      0
#> 37   -1.6834133330      0
#> 38   -1.4291832133      0
#> 39   -1.9820566769      0
#> 40   -0.4931104273      0
#> 41    1.4400915084      0
#> 42    1.1478103254      0
#> 43    0.4271144243      0
#> 44   -0.5870214477      0
#> 45   -0.1223026752      0
#> 46   -1.9464882388      0
#> 47    2.2623843679      0
#> 48   -0.3803177532      0
#> 49    0.9115241790      0
#> 50    0.1190581380      0
#> 51    0.3519171193      0
#> 52   -0.4662320643      0
#> 53   -0.0417615072      0
#> 54    0.5343514616      0
#> 55   -0.2660683844      0
#> 56   -0.3367377364      0
#> 57   -1.3593430903      0
#> 58    1.9559293001      0
#> 59   -0.5315264694      0
#> 60    1.1579323418      0
#> 61   -0.2309626566      0
#> 62   -0.8616974473      0
#> 63   -1.6018348732      0
#> 64    0.2896419111      0
#> 65   -0.4265044020      0
#> 66    1.6732261856      0
#> 67    0.2525366756      0
#> 68    0.4646232417      0
#> 69    0.0437393886      0
#> 70    0.0467884667      0
#> 71   -0.6756980687      0
#> 72   -0.1013874516      0
#> 73    0.0117372018      0
#> 74    0.0528058373      0
#> 75    1.0808970940      0
#> 76   -0.7463022691      0
#> 77   -0.0217686648      0
#> 78    0.7000538494      0
#> 79   -0.1277582298      0
#> 80    1.9374750337      0
#> 81   -0.9972345269      0
#> 82    0.8403049365      0
#> 83   -0.3265497503      0
#> 84    0.9922546789      0
#> 85    0.4647273462      0
#> 86    0.9648985066      0
#> 87   -2.0249314722      0
#> 88    0.0546696425      0
#> 89   -0.4597373521      0
#> 90    0.2146062439      0
#> 91    0.0399484502      0
#> 92   -0.7145904502      0
#> 93   -0.5055544526      0
#> 94    1.3010691928      0
#> 95   -0.2594448903      0
#> 96   -1.0428378833      0
#> 97    0.2638588220      0
#> 98    1.5250837246      0
#> 99    2.0125696612      0
#> 100  -0.9745303851      0
#> 101   0.5424087267      0
#> 102   1.0090320492      0
#> 103  -0.2328799515      0
#> 104   2.4619186574      0
#> 105   1.1788411309      0
#> 106   0.1310842664      0
#> 107   0.9315844690      0
#> 108  -1.7416955639      0
#> 109   0.5829206167      0
#> 110   1.8920582431      0
#> 111  -0.1746085106      0
#> 112  -0.9157073463      0
#> 113   0.6730526196      0
#> 114   0.7989978484      0
#> 115   0.2242131959      0
#> 116  -1.6497938600      0
#> 117  -0.8541898798      0
#> 118   1.6052074833      0
#> 119   0.1894565834      0
#> 120  -1.6596971521      0
#> 121   0.8658492322      0
#> 122   0.3145712568      0
#> 123  -0.5763932503      0
#> 124  -0.6576200336      0
#> 125   1.0342967231      0
#> 126   0.8986888554      0
#> 127  -1.1693364321      0
#> 128   0.0394722910      0
#> 129  -0.0429359607      0
#> 130  -0.0781346584      0
#> 131  -0.1384637887      0
#> 132  -0.3776263849      0
#> 133  -0.3500076406      0
#> 134   1.3581810935      0
#> 135  -0.7916336561      0
#> 136   1.0792406351      0
#> 137   2.9460503658      0
#> 138   1.7999795654      0
#> 139  -0.5650369835      0
#> 140   0.4899073437      0
#> 141   0.0160575753      0
#> 142   0.0958320963      0
#> 143   1.6224022872      0
#> 144   0.2312302108      0
#> 145  -0.3458672950      0
#> 146   0.7262925431      0
#> 147  -0.2687088550      0
#> 148   0.0397201354      0
#> 149   0.9176429093      0
#> 150  -0.2504870985      0
#> 151  -0.9030746419      0
#> 152   0.7005600791      0
#> 153  -0.7992883025      0
#> 154   1.2652754254      0
#> 155   1.3369884266      0
#> 156  -0.4385325017      0
#> 157   0.5803912815      0
#> 158  -2.0391662077      0
#> 159   0.5662188649      0
#> 160  -0.2358351443      0
#> 161   1.3078068276      0
#> 162  -0.4114866425      0
#> 163  -0.6333470857      0
#> 164  -0.4998506934      0
#> 165  -1.1525194618      0
#> 166   0.8895905704      0
#> 167   0.3108546584      0
#> 168  -1.1321718655      0
#> 169  -1.6350914170      0
#> 170  -0.5828214177      0
#> 171   0.0499852989      0
#> 172  -0.5926655012      0
#> 173  -1.4373667804      0
#> 174   0.2481660613      0
#> 175   0.7564900953      0
#> 176  -0.6580484867      0
#> 177  -0.2160181060      0
#> 178  -0.8819435321      0
#> 179  -1.2916570940      0
#> 180  -1.5706135570      0
#> 181   1.1964606339      0
#> 182  -0.3327392221      0
#> 183   1.1444734801      0
#> 184   1.9136468416      0
#> 185  -0.4466414459      0
#> 186   1.1794772988      0
#> 187   0.6517880591      0
#> 188  -0.6194571711      0
#> 189   0.1093626686      0
#> 190  -0.9066588534      0
#> 191   1.0309151829      0
#> 192   0.8351740354      0
#> 193   0.8100135801      0
#> 194  -1.4327179537      0
#> 195  -0.2772109146      0
#> 196  -0.3410147233      0
#> 197   2.5079071736      0
#> 198  -0.3035335294      0
#> 199  -1.1147453581      0
#> 200  -0.1995693264      0
#> 201  -0.0820879801      0
#> 202  -0.4709724402      0
#> 203  -0.2241192440      0
#> 204  -0.6276647940      0
#> 205   0.4887812496      0
#> 206  -0.0276280652      0
#> 207   0.0328437011      0
#> 208  -0.2040890134      0
#> 209  -1.8084670004      0
#> 210   1.3047674653      0
#> 211  -0.8175952645      0
#> 212   1.5267800009      0
#> 213   1.2762903501      0
#> 214   0.4967961960      0
#> 215   0.0711120362      0
#> 216   1.2814114460      0
#> 217   0.2172968953      0
#> 218   0.2733326369      0
#> 219   0.1041692193      0
#> 220  -0.5455316439      0
#> 221  -0.2547987189      0
#> 222  -1.5443171771      0
#> 223  -1.0127077218      0
#> 224   0.7442329226      0
#> 225   0.3820756299      0
#> 226   0.5034128063      0
#> 227   0.1628193630      0
#> 228  -0.0171916618      0
#> 229   0.1303645445      0
#> 230   0.9009441254      0
#> 231  -1.4106133137      0
#> 232   0.0598725947      0
#> 233  -2.1067333877      0
#> 234   0.2866157611      0
#> 235   1.4023530607      0
#> 236  -0.0537820831      0
#> 237   0.4186977441      0
#> 238   0.0080542407      0
#> 239  -2.1873663253      0
#> 240   0.6192937704      0
#> 241  -0.6737385357      0
#> 242   1.5554835957      0
#> 243   0.0323322570      0
#> 244  -0.9226353636      0
#> 245  -0.4766315547      0
#> 246   1.3099606828      0
#> 247  -0.6451101080      0
#> 248  -0.9676586018      0
#> 249  -0.1917607146      0
#> 250  -1.7541818900      0
#> 251   1.6144830513      0
#> 252   1.1909704085      0
#> 253  -1.5032329324      0
#> 254  -0.2600614823      0
#> 255  -0.6156154282      0
#> 256  -1.4197082797      0
#> 257  -0.3322546457      0
#> 258  -0.9990099223      0
#> 259   0.3847228802      0
#> 260   0.2701523722      0
#> 261   0.0403607517      0
#> 262  -0.4376717318      0
#> 263   0.2987111172      0
#> 264  -0.6716614703      0
#> 265  -0.9643685823      0
#> 266   1.7591502867      0
#> 267  -0.2788183259      0
#> 268   0.7642158743      0
#> 269  -1.0565080954      0
#> 270  -2.2669127726      0
#> 271  -0.2245168877      0
#> 272   0.5736945133      0
#> 273   0.1691795362      0
#> 274  -0.4521136817      0
#> 275   0.0939036301      0
#> 276  -0.0079864351      0
#> 277   0.0183442006      0
#> 278  -1.4336917768      0
#> 279   0.0308151554      0
#> 280  -1.2570573604      0
#> 281   1.2741544128      0
#> 282  -0.3567676938      0
#> 283   0.0837200693      0
#> 284   0.1328384148      0
#> 285  -0.4324687317      0
#> 286   0.1236853774      0
#> 287   0.1664105798      0
#> 288   0.6339989365      0
#> 289   1.2824550962      0
#> 290   1.2297348398      0
#> 291  -2.2103628746      0
#> 292  -0.5958975801      0
#> 293   1.3062138543      0
#> 294  -0.3629410729      0
#> 295   0.7593216418      0
#> 296  -1.0963765417      0
#> 297  -1.3926169947      0
#> 298   1.5674809374      0
#> 299  -0.5520565521      0
#> 300  -0.5187067057      0
#> 301  -0.3335603233      0
#> 302   1.3989610661      0
#> 303   1.2934059681      0
#> 304  -1.0664164973      0
#> 305   0.2886472378      0
#> 306  -0.1762630395      0
#> 307   0.2262190549      0
#> 308   0.5030379898      0
#> 309   1.2666677947      0
#> 310  -0.0046702637      0
#> 311   1.5188733222      0
#> 312  -0.6920902469      0
#> 313  -0.3445065372      0
#> 314   2.3215365881      0
#> 315   0.1898356427      0
#> 316   0.0513681551      0
#> 317   1.2842566597      0
#> 318  -1.4722723434      0
#> 319  -1.1075357942      0
#> 320   0.1236414502      0
#> 321  -0.8782008789      0
#> 322  -0.3148472599      0
#> 323  -0.0543476438      0
#> 324  -0.7188071571      0
#> 325  -0.4666287099      0
#> 326  -2.1260118327      0
#> 327   0.2587410935      0
#> 328   0.0150600136      0
#> 329  -1.2983351129      0
#> 330  -0.8240480146      0
#> 331  -0.5992943034      0
#> 332   0.3406411611      0
#> 333  -2.1701193651      0
#> 334  -0.5117214605      0
#> 335  -0.4867376918      0
#> 336   0.3347235872      0
#> 337   0.9127777991      0
#> 338  -0.4889583381      0
#> 339  -0.8764601518      0
#> 340   1.0265644214      0
#> 341   0.4258575526      0
#> 342   1.7749151773      0
#> 343   0.3027206885      0
#> 344   0.4388952514      0
#> 345   2.0255328683      0
#> 346   0.6706948705      0
#> 347   0.9499762700      0
#> 348   0.3846450506      0
#> 349  -0.6568408288      0
#> 350   0.7119679784      0
#> 351  -0.0352929089      0
#> 352   0.4592142625      0
#> 353   1.1260166858      0
#> 354  -0.5114946026      0
#> 355   0.4249109173      0
#> 356   0.0554646546      0
#> 357  -0.0550450534      0
#> 358   0.2645581689      0
#> 359  -1.8156962783      0
#> 360   0.9553081785      0
#> 361   1.5903264502      0
#> 362  -0.2699529293      0
#> 363  -1.2297047869      0
#> 364  -0.5029222965      0
#> 365  -0.1899033555      0
#> 366   0.9431334986      0
#> 367   0.6112276577      0
#> 368   0.1234225661      0
#> 369   0.6743287521      0
#> 370   1.8750410611      0
#> 371  -1.1486031007      0
#> 372   0.2524654796      0
#> 373   0.6081030122      0
#> 374   0.6896155885      0
#> 375   1.7827463583      0
#> 376   1.3758255492      0
#> 377   0.5609472033      0
#> 378  -0.8483473317      0
#> 379   1.4884622863      0
#> 380   0.2693500171      0
#> 381  -0.2988234588      0
#> 382  -1.1117706980      0
#> 383   0.8942218367      0
#> 384   0.4390923449      0
#> 385  -0.5903040810      0
#> 386  -0.6908142291      0
#> 387   2.8131425435      0
#> 388   0.5271369916      0
#> 389   1.9517955404      0
#> 390   1.3593402068      0
#> 391   0.5312344129      0
#> 392   0.1298520453      0
#> 393  -0.2213606345      0
#> 394   0.7576106832      0
#> 395   0.3277030422      0
#> 396   1.1864139702      0
#> 397   0.2808304208      0
#> 398  -0.7290552817      0
#> 399   1.2427716549      0
#> 400  -0.7299990017      0
#> 401  -1.0320374915      0
#> 402   2.0041170022      0
#> 403   0.2861745814      0
#> 404  -0.2521560700      0
#> 405   1.9066456666      0
#> 406  -0.0743738951      0
#> 407   1.2698433799      0
#> 408   1.3598054096      0
#> 409  -0.9868941414      0
#> 410  -0.0745024517      0
#> 411  -0.3792729780      0
#> 412   1.8760669522      0
#> 413  -0.0813691908      0
#> 414  -0.7985919841      0
#> 415  -0.0756806133      0
#> 416   1.3520463189      0
#> 417  -1.3141815400      0
#> 418  -0.7385297017      0
#> 419   0.4689823796      0
#> 420  -0.1738530788      0
#> 421   0.6436208807      0
#> 422  -0.1410838019      0
#> 423   0.4680889270      0
#> 424  -0.0868452106      0
#> 425  -0.7973455109      0
#> 426  -0.4708998846      0
#> 427  -0.4718922957      0
#> 428   0.3171479174      0
#> 429  -0.0263166194      0
#> 430   0.6490094912      0
#> 431   1.2088840877      0
#> 432  -0.8104159693      0
#> 433   0.1893194043      0
#> 434  -1.5858934003      0
#> 435   0.4024846745      0
#> 436  -1.1098193094      0
#> 437  -1.0140208095      0
#> 438   0.4001230200      0
#> 439   0.2478806102      0
#> 440   1.0679193791      0
#> 441   0.9686232091      0
#> 442   0.2399661985      0
#> 443  -1.4146754645      0
#> 444  -2.1717312940      0
#> 445  -0.0698460831      0
#> 446  -0.1904734990      0
#> 447  -2.7382989385      0
#> 448  -0.7495794657      0
#> 449   0.1308721423      0
#> 450  -0.7925334535      0
#> 451  -0.9050539206      0
#> 452   0.4366527240      0
#> 453  -0.6514596748      0
#> 454   0.2051453974      0
#> 455   0.3990405532      0
#> 456  -0.5648381403      0
#> 457   0.0156150281      0
#> 458  -0.8058888297      0
#> 459   0.8260271724      0
#> 460  -0.7949844403      0
#> 461   0.1992918690      0
#> 462   1.2757219584      0
#> 463  -0.1655222246      0
#> 464  -0.6877938842      0
#> 465   1.1756637134      0
#> 466   1.8472034478      0
#> 467   0.7523765071      0
#> 468  -0.2453342885      0
#> 469  -1.4339415037      0
#> 470   0.3796360916      0
#> 471   0.2036041309      0
#> 472   1.5755763372      0
#> 473  -0.3576457536      0
#> 474  -2.3536517532      0
#> 475   0.4438526156      0
#> 476   0.6680844944      0
#> 477   1.3733475773      0
#> 478  -1.0862483344      0
#> 479  -1.2268433772      0
#> 480   0.5018184692      0
#> 481   0.0397736375      0
#> 482   2.9409919376      0
#> 483   0.7297251659      0
#> 484  -0.7325819685      0
#> 485  -0.8836738953      0
#> 486   0.6754414353      0
#> 487  -0.8619956967      0
#> 488   0.1718283980      0
#> 489  -1.5016718043      0
#> 490   1.4589442792      0
#> 491  -0.1032257128      0
#> 492  -1.8509807515      0
#> 493   0.3194058385      0
#> 494  -0.7747408445      0
#> 495  -0.3724430442      0
#> 496  -0.9399598524      0
#> 497  -0.4203468397      0
#> 498   0.0343035922      0
#> 499  -0.3855727474      0
#> 500   0.4785478591      0
#> 501   0.1675837633      0
#> 502   0.9666768478      0
#> 503  -0.1575690999      0
#> 504   1.5013196587      0
#> 505   0.4394577144      0
#> 506  -0.4075312365      0
#> 507   2.2955855598      0
#> 508  -0.0588472359      0
#> 509   1.7425866884      0
#> 510   0.7910093490      0
#> 511   1.4369774005      0
#> 512   0.0833344742      0
#> 513   1.8043986048      0
#> 514  -1.8185200314      0
#> 515   0.5498682608      0
#> 516  -1.1069104237      0
#> 517  -0.2737862336      0
#> 518  -2.2301131190      0
#> 519   0.7748657500      0
#> 520   0.7876412787      0
#> 521  -0.8643498555      0
#> 522  -1.3712061516      0
#> 523  -0.2147190333      0
#> 524  -0.9124382038      0
#> 525   1.5105127534      0
#> 526   0.6425057587      0
#> 527   2.1562612258      0
#> 528   0.8346875521      0
#> 529   0.5717621885      0
#> 530  -0.6140851395      0
#> 531  -1.0772160917      0
#> 532   0.8900786643      0
#> 533  -0.3108134107      0
#> 534   0.8254450141      0
#> 535  -0.7239757240      0
#> 536  -1.4775335021      0
#> 537   1.1199033817      0
#> 538   0.2593503519      0
#> 539   1.4585357704      0
#> 540   0.9250118593      0
#> 541   0.1705514094      0
#> 542  -0.6671833109      0
#> 543   0.3792065593      0
#> 544   0.6721598480      0
#> 545  -1.0900188104      0
#> 546  -1.7278955570      0
#> 547   2.2708732700      0
#> 548  -2.0445629836      0
#> 549   1.5678084184      0
#> 550   3.0449281263      0
#> 551   0.7915149267      0
#> 552   0.7363196373      0
#> 553   2.5979956610      0
#> 554   1.7824443349      0
#> 555   0.8857774891      0
#> 556  -0.9787609897      0
#> 557  -0.3332364398      0
#> 558  -1.2549186943      0
#> 559   0.8796861498      0
#> 560   0.4043484291      0
#> 561   1.6786275003      0
#> 562   0.7537985854      0
#> 563   0.7361955219      0
#> 564  -0.2971467375      0
#> 565  -0.3749354476      0
#> 566   0.1032686022      0
#> 567  -1.0353631998      0
#> 568  -1.9142178725      0
#> 569  -0.5758150156      0
#> 570  -0.3549904912      0
#> 571   0.2637322876      0
#> 572  -0.5708277287      0
#> 573   0.5947253811      0
#> 574  -1.4782808603      0
#> 575   0.9732378768      0
#> 576   0.7934875069      0
#> 577   1.8580550930      0
#> 578  -0.2136247669      0
#> 579  -0.8251458055      0
#> 580   1.5876741388      0
#> 581  -0.7363109522      0
#> 582  -1.5808681877      0
#> 583   0.3674818126      0
#> 584   0.0319857646      0
#> 585   0.4444907155      0
#> 586   1.5351444571      0
#> 587   0.0330669135      0
#> 588  -0.7784315632      0
#> 589  -0.8784232756      0
#> 590   0.4350779607      0
#> 591  -0.0052682225      0
#> 592   2.1289831278      0
#> 593   0.7555052744      0
#> 594  -0.9335601259      0
#> 595   0.0617767767      0
#> 596   1.0785827907      0
#> 597  -2.0877381640      0
#> 598   1.8826618117      0
#> 599  -1.2789370381      0
#> 600  -1.0726863265      0
#> 601   0.1872363545      0
#> 602   0.2290322398      0
#> 603   0.7763029760      0
#> 604   0.2618268559      0
#> 605   0.0234349945      0
#> 606   0.8965946132      0
#> 607   0.1731950839      0
#> 608   1.2259044424      0
#> 609   0.0648820892      0
#> 610   0.2889685367      0
#> 611   0.7329970201      0
#> 612  -1.2646925753      0
#> 613   1.5255762137      0
#> 614  -0.1373275720      0
#> 615  -0.2398132936      0
#> 616  -0.2916685893      0
#> 617   1.1942794438      0
#> 618   0.9593673148      0
#> 619  -0.6979718459      0
#> 620   0.5076378099      0
#> 621  -0.1867474906      0
#> 622   0.1510507780      0
#> 623  -0.2953938865      0
#> 624   0.3905089233      0
#> 625  -0.2809581001      0
#> 626  -0.7727199607      0
#> 627   0.0082996988      0
#> 628  -0.4584978224      0
#> 629   0.6638369640      0
#> 630   0.8593107058      0
#> 631   0.6859592723      0
#> 632  -0.7335206684      0
#> 633   1.4519199510      0
#> 634   0.0106587247      0
#> 635  -0.2403593181      0
#> 636   0.8471752655      0
#> 637   0.3818717983      0
#> 638   1.1341951876      0
#> 639  -0.1024879726      0
#> 640  -0.6673748876      0
#> 641   0.8290081790      0
#> 642  -0.8723851739      0
#> 643   0.7655942262      0
#> 644   0.4610625991      0
#> 645  -0.1245832046      0
#> 646  -1.2015169628      0
#> 647  -3.4278438504      0
#> 648  -0.4590655260      0
#> 649   0.7553213338      0
#> 650   0.3837618087      0
#> 651   0.8589907275      0
#> 652  -1.8941857202      0
#> 653   1.1724920853      0
#> 654   1.5865695190      0
#> 655   0.5425162249      0
#> 656  -0.2514466477      0
#> 657  -1.8340101987      0
#> 658   0.1883327989      0
#> 659  -0.9575584578      0
#> 660  -0.5244821590      0
#> 661   1.6606766671      0
#> 662   0.6989787439      0
#> 663  -1.1336006959      0
#> 664  -1.0843622692      0
#> 665  -0.3589309819      0
#> 666   0.6067739442      0
#> 667  -1.0706985076      0
#> 668  -0.7401378468      0
#> 669  -1.1278719372      0
#> 670   1.3474354339      0
#> 671   0.3597838387      0
#> 672  -0.0678559385      0
#> 673   0.0006838311      0
#> 674   0.8289935236      0
#> 675  -0.4161026188      0
#> 676   0.2093962523      0
#> 677   0.7865674936      0
#> 678  -0.3463047943      0
#> 679   0.6751679080      0
#> 680   2.3874493337      0
#> 681   0.7681222524      0
#> 682  -1.2909140315      0
#> 683   0.3688597697      0
#> 684  -0.8092401068      0
#> 685  -0.4411210941      0
#> 686   1.1890351729      0
#> 687  -0.3506548360      0
#> 688  -0.4644440924      0
#> 689  -0.2305843795      0
#> 690   0.0705575920      0
#> 691   0.9806635962      0
#> 692  -0.2332246022      0
#> 693   0.7986873943      0
#> 694  -1.3788977354      0
#> 695   0.8686884051      0
#> 696  -1.4646395533      0
#> 697  -1.0652227040      0
#> 698   0.5991937805      0
#> 699  -0.3953105438      0
#> 700  -1.0033677861      0
#> 701  -0.0929015809      0
#> 702   1.0968826345      0
#> 703  -1.6009018447      0
#> 704   0.9007475532      0
#> 705  -0.5046165628      0
#> 706   0.5557824185      0
#> 707  -1.0386874850      0
#> 708   0.0192558186      0
#> 709  -0.1075002573      0
#> 710  -1.2868768165      0
#> 711   0.2434391747      0
#> 712  -0.6382283024      0
#> 713  -0.1768986525      0
#> 714   0.7407966711      0
#> 715  -0.5263951931      0
#> 716   0.4959930513      0
#> 717  -0.1412951941      0
#> 718   0.7927512345      0
#> 719  -1.0625190434      0
#> 720   1.9886141492      0
#> 721   0.6022454210      0
#> 722  -1.4147107808      0
#> 723  -0.5922221073      0
#> 724  -1.3995041703      0
#> 725  -0.8110691773      0
#> 726  -0.7897988625      0
#> 727  -0.2514503499      0
#> 728   0.7641514625      0
#> 729   1.0340210858      0
#> 730   0.1135129622      0
#> 731   1.7570006229      0
#> 732   1.1161064492      0
#> 733  -0.0232221422      0
#> 734   1.8710586875      0
#> 735  -0.2998360059      0
#> 736  -0.0519048955      0
#> 737  -0.2699623443      0
#> 738  -0.4164221346      0
#> 739  -1.5575507356      0
#> 740  -1.1179939848      0
#> 741   0.9327186131      0
#> 742  -1.4817519728      0
#> 743   1.0667888078      0
#> 744   1.6331065908      0
#> 745  -0.8561450150      0
#> 746  -2.0229340958      0
#> 747  -0.3843779549      0
#> 748   0.7724681208      0
#> 749  -0.3093355399      0
#> 750   2.1028730542      0
#> 751   0.7386444527      0
#> 752  -0.2931754457      0
#> 753  -1.6606919544      0
#> 754  -0.9223976093      0
#> 755   0.7146688449      0
#> 756   0.3572826543      0
#> 757   1.0391848916      0
#> 758   1.1957596555      0
#> 759  -0.0129948080      0
#> 760  -2.0862950489      0
#> 761   0.2311659853      0
#> 762  -2.5663849791      0
#> 763   1.5814560836      0
#> 764   0.5580553519      0
#> 765   0.3277079404      0
#> 766  -0.7413642978      0
#> 767   1.5820414422      0
#> 768   0.8959441492      0
#> 769  -0.1695705987      0
#> 770  -1.0154855806      0
#> 771   0.4007641269      0
#> 772  -0.5265120136      0
#> 773  -1.9438530649      0
#> 774   1.7249309195      0
#> 775   1.5750714578      0
#> 776  -1.1686990581      0
#> 777   0.6014013710      0
#> 778   0.5706876612      0
#> 779   0.1074936474      0
#> 780  -0.7379104142      0
#> 781  -0.1254882504      0
#> 782   0.3683483495      0
#> 783  -0.1093133792      0
#> 784   1.2182240413      0
#> 785   1.6677212109      0
#> 786   0.7698491664      0
#> 787   0.3119965978      0
#> 788   0.4346713236      0
#> 789   1.1288091204      0
#> 790  -2.5126912099      0
#> 791  -0.6791505852      0
#> 792   0.6096783909      0
#> 793   1.0611717063      0
#> 794  -1.1486814955      0
#> 795   1.0978802005      0
#> 796  -0.7987877414      0
#> 797  -0.5566363057      0
#> 798  -0.8288804968      0
#> 799   1.0773430426      0
#> 800  -0.5080624322      0
#> 801  -1.0601462172      0
#> 802  -1.1932911103      0
#> 803  -0.0812787591      0
#> 804   0.2615370851      0
#> 805  -0.2657282498      0
#> 806  -0.7774773399      0
#> 807  -1.0220937865      0
#> 808   1.0948297022      0
#> 809   0.1318183150      0
#> 810   0.4113411904      0
#> 811   1.7816003139      0
#> 812   0.6870051668      0
#> 813   0.5271418630      0
#> 814   0.0579278142      0
#> 815   0.3741026849      0
#> 816   0.1561327856      0
#> 817   1.2907517167      0
#> 818   0.9350477544      0
#> 819   0.9073758412      0
#> 820   0.1243254770      0
#> 821  -0.5787743231      0
#> 822  -0.1439541431      0
#> 823  -0.5907373764      0
#> 824  -0.4975940650      0
#> 825   0.2633715702      0
#> 826   0.1156082549      0
#> 827   0.6141493885      0
#> 828   0.0801914659      0
#> 829   1.1610652693      0
#> 830   1.3826311340      0
#> 831  -1.2350820436      0
#> 832   0.1274975504      0
#> 833  -1.2478317000      0
#> 834   0.2262967764      0
#> 835   0.0426952899      0
#> 836   1.1900420156      0
#> 837  -0.8417769513      0
#> 838   0.0555261252      0
#> 839   0.7451835040      0
#> 840   0.3629393099      0
#> 841  -0.3825429153      0
#> 842   0.9103136269      0
#> 843  -0.8352351253      0
#> 844   0.9101031991      0
#> 845  -1.6996140924      0
#> 846   0.3765185823      0
#> 847   0.4794332509      0
#> 848  -1.6282841282      0
#> 849  -0.1791952180      0
#> 850  -1.3076807421      0
#> 851  -0.0362713001      0
#> 852  -0.5820217437      0
#> 853  -0.2793276776      0
#> 854  -0.1719432921      0
#> 855  -1.3442342575      0
#> 856  -1.0122569212      0
#> 857   0.1655443080      0
#> 858  -0.0559416181      0
#> 859  -0.4222344651      0
#> 860  -2.3985700037      0
#> 861  -0.2219100578      0
#> 862  -0.4110486696      0
#> 863   1.3392733233      0
#> 864  -0.3493972520      0
#> 865   0.0686435785      0
#> 866   0.7537422913      0
#> 867   0.4544721380      0
#> 868  -0.8956570983      0
#> 869  -0.6609711661      0
#> 870   1.3097434918      0
#> 871   1.0672249427      0
#> 872  -0.5789159629      0
#> 873   2.0848036134      0
#> 874  -0.1690319444      0
#> 875   1.0484561421      0
#> 876  -0.6780099592      0
#> 877   0.3784208158      0
#> 878  -0.4672910602      0
#> 879   0.3995704996      0
#> 880  -0.3189971112      0
#> 881  -0.3908163233      0
#> 882   1.1669595907      0
#> 883   0.7884191518      0
#> 884  -0.5940689173      0
#> 885   0.2295293657      0
#> 886   0.6355526616      0
#> 887   0.6867608527      0
#> 888  -0.8775320209      0
#> 889   0.5208431276      0
#> 890   2.0380673557      0
#> 891  -1.1933364818      0
#> 892   0.9522827374      0
#> 893  -1.9661360713      0
#> 894  -0.1022242376      0
#> 895   0.1320939825      0
#> 896  -0.0079698331      0
#> 897   0.1447483973      0
#> 898   0.6833141347      0
#> 899  -0.6882383546      0
#> 900   1.1227158713      0
#> 901  -1.1515791794      0
#> 902   1.7708459382      0
#> 903   1.0331605665      1
#> 904   1.6323853180      1
#> 905  -1.9469618189      1
#> 906  -0.2651979058      1
#> 907  -0.2116597882      1
#> 908   0.6056926535      1
#> 909   1.2244892350      1
#> 910   0.3356224506      1
#> 911   1.8704145326      1
#> 912   0.7335336008      1
#> 913   0.0239037827      1
#> 914  -0.7450697800      1
#> 915   1.8393832178      1
#> 916   0.0428856538      1
#> 917   0.6766232257      1
#> 918  -0.0600563525      1
#> 919  -0.6900046666      1
#> 920   1.3320333802      1
#> 921   0.6635079883      1
#> 922  -1.0206622530      1
#> 923   0.9683050259      1
#> 924   1.8142424519      1
#> 925   2.0301176383      1
#> 926  -0.2046759270      1
#> 927   1.5266990592      1
#> 928   1.7374496398      1
#> 929   1.6630438633      1
#> 930   0.2066816914      1
#> 931   0.9847838122      1
#> 932   0.0453273169      1
#> 933   0.1537909718      1
#> 934   1.4054937202      1
#> 935   0.3111200618      1
#> 936   1.2746917616      1
#> 937   1.4302628605      1
#> 938   1.3047489549      1
#> 939   0.7454710971      1
#> 940   1.6322572240      1
#> 941   0.1400844508      1
#> 942  -0.5898182190      1
#> 943   2.4143721022      1
#> 944  -0.3464611572      1
#> 945   1.2942869239      1
#> 946   0.4206190654      1
#> 947   1.3019043443      1
#> 948   2.9202003383      1
#> 949  -0.2404677929      1
#> 950  -0.0520841183      1
#> 951   0.1586743644      1
#> 952   0.5658525570      1
#> 953   0.8585616682      1
#> 954  -0.9479469687      1
#> 955   0.0976160166      1
#> 956   1.5059656140      1
#> 957   0.0473562312      1
#> 958   1.5555627841      1
#> 959   2.0343699711      1
#> 960   2.5808843949      1
#> 961   1.7367081837      1
#> 962   0.0225476329      1
#> 963   1.4913111939      1
#> 964   0.7981369838      1
#> 965   1.0863269751      1
#> 966  -0.6077579334      1
#> 967   0.6135021812      1
#> 968   0.4277513118      1
#> 969   0.7536323928      1
#> 970  -0.2683374984      1
#> 971   0.7603257883      1
#> 972   1.6154160524      1
#> 973  -0.3491103026      1
#> 974   2.5741538476      1
#> 975  -0.2608835191      1
#> 976   1.2819848762      1
#> 977  -1.0717579196      1
#> 978  -0.0078945210      1
#> 979  -0.4883570526      1
#> 980  -0.0046497121      1
#> 981  -0.8559478712      1
#> 982   0.2562997965      1
#> 983   0.5801480984      1
#> 984   0.5850538994      1
#> 985  -0.8264420728      1
#> 986   1.4397703914      1
#> 987   0.3697803112      1
#> 988   0.5152841865      1
#> 989   1.3479926005      1
#> 990   2.7305819603      1
#> 991  -0.2826579237      1
#> 992   1.3468686697      1
#> 993   0.9582838602      1
#> 994   1.8089107956      1
#> 995   1.0984040952      1
#> 996  -0.0291618133      1
#> 997   0.4953772352      1
#> 998  -0.5516171531      1
#> 999   2.0988423165      1
#> 1000  1.3251303351      1