Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                 x actual
#> 1    -0.617461930      0
#> 2     0.569870616      0
#> 3     0.035603183      0
#> 4     0.177463123      0
#> 5     1.146876273      0
#> 6     0.715607894      0
#> 7    -0.536672119      0
#> 8     1.657865319      0
#> 9     0.224683236      0
#> 10   -0.230227626      0
#> 11   -0.252366584      0
#> 12    1.849244116      0
#> 13    0.170529657      0
#> 14   -0.950478585      0
#> 15   -0.750147728      0
#> 16    0.429271826      0
#> 17   -0.240573488      0
#> 18    0.048679850      0
#> 19   -0.610360015      0
#> 20    0.756738949      0
#> 21   -0.081809457      0
#> 22   -0.527199449      0
#> 23   -0.630139397      0
#> 24   -0.466149604      0
#> 25   -0.241976236      0
#> 26    0.014042533      0
#> 27   -1.422289566      0
#> 28    0.615853774      0
#> 29    0.212073591      0
#> 30    0.141040922      0
#> 31    0.648030281      0
#> 32   -0.465142247      0
#> 33   -1.555816566      0
#> 34    0.010616493      0
#> 35   -0.778073821      0
#> 36    0.284292345      0
#> 37    0.157833087      0
#> 38   -2.135788837      0
#> 39    0.720858625      0
#> 40    0.469421211      0
#> 41    0.989725440      0
#> 42   -0.841472939      0
#> 43   -0.660649513      0
#> 44   -0.194238163      0
#> 45    2.148712336      0
#> 46   -0.457014962      0
#> 47    0.954446308      0
#> 48    0.118205509      0
#> 49   -2.199660011      0
#> 50    0.915048276      0
#> 51   -0.125212127      0
#> 52    1.411598347      0
#> 53   -2.837259503      0
#> 54   -1.214185345      0
#> 55   -2.258659008      0
#> 56   -0.592410698      0
#> 57   -1.173708007      0
#> 58   -0.137822952      0
#> 59    0.040889197      0
#> 60   -1.595668846      0
#> 61   -1.162916511      0
#> 62    1.070059276      0
#> 63   -0.022479167      0
#> 64    0.506948274      0
#> 65   -0.943546951      0
#> 66   -0.541246706      0
#> 67    0.110917533      0
#> 68    1.238949734      0
#> 69   -0.352683473      0
#> 70   -0.530064443      0
#> 71    0.547492011      0
#> 72    0.867766975      0
#> 73   -0.213588079      0
#> 74    1.628222857      0
#> 75    0.643877746      0
#> 76   -1.289889832      0
#> 77   -0.596853006      0
#> 78    1.409747704      0
#> 79   -0.807401752      0
#> 80   -0.837144875      0
#> 81   -0.306363365      0
#> 82    1.724452005      0
#> 83   -0.549957780      0
#> 84   -0.032001440      0
#> 85    0.917193949      0
#> 86    0.254322269      0
#> 87    0.530295853      0
#> 88   -0.230884038      0
#> 89    1.064024868      0
#> 90   -1.820010422      0
#> 91    0.493961610      0
#> 92   -0.570506310      0
#> 93   -0.919806190      0
#> 94   -0.668254916      0
#> 95    0.828912695      0
#> 96    0.323819606      0
#> 97   -0.363556900      0
#> 98   -0.940716317      0
#> 99    0.353096987      0
#> 100   0.738231787      0
#> 101   0.711740167      0
#> 102   0.117246915      0
#> 103   1.214874428      0
#> 104  -1.343024505      0
#> 105  -0.122946406      0
#> 106  -0.024549988      0
#> 107   0.221567115      0
#> 108   1.107351690      0
#> 109  -0.131959704      0
#> 110  -0.506279339      0
#> 111   0.299060914      0
#> 112  -0.721948143      0
#> 113  -0.567982339      0
#> 114   2.761999542      0
#> 115   0.368984812      0
#> 116  -1.144279357      0
#> 117   0.393405804      0
#> 118  -0.211149114      0
#> 119  -0.679600493      0
#> 120   1.848971745      0
#> 121  -0.700277660      0
#> 122  -0.068365572      0
#> 123   0.142756737      0
#> 124   2.068224892      0
#> 125   1.820664564      0
#> 126   0.010991790      0
#> 127   0.194184285      0
#> 128  -1.068155006      0
#> 129  -0.110675747      0
#> 130   0.564664091      0
#> 131  -0.938204277      0
#> 132   0.894054052      0
#> 133  -0.521313088      0
#> 134   0.406442339      0
#> 135   0.314228783      0
#> 136  -1.534758880      0
#> 137  -0.701305806      0
#> 138  -0.909092731      0
#> 139   1.187976574      0
#> 140  -0.135601675      0
#> 141   0.586960944      0
#> 142   0.838039363      0
#> 143  -0.859618455      0
#> 144  -2.447952440      0
#> 145  -0.484347083      0
#> 146  -1.225429847      0
#> 147  -1.573073678      0
#> 148  -1.729418653      0
#> 149   0.645563576      0
#> 150  -0.088633549      0
#> 151   0.604413892      0
#> 152   0.246739449      0
#> 153  -0.316888317      0
#> 154   0.240150095      0
#> 155   0.142876571      0
#> 156  -0.479233547      0
#> 157   0.400456379      0
#> 158  -0.683750943      0
#> 159  -1.251800481      0
#> 160   0.896325778      0
#> 161   0.549598987      0
#> 162  -0.400770277      0
#> 163  -0.070478258      0
#> 164   0.677893057      0
#> 165   1.802244199      0
#> 166  -0.942272762      0
#> 167  -0.923904459      0
#> 168   0.134374451      0
#> 169  -0.629671110      0
#> 170   1.013196183      0
#> 171  -0.194243591      0
#> 172   1.077907079      0
#> 173   0.628045636      0
#> 174   1.352814405      0
#> 175  -0.764683437      0
#> 176  -1.533863789      0
#> 177   0.471607752      0
#> 178  -0.052705879      0
#> 179  -0.533712568      0
#> 180  -0.196719906      0
#> 181   0.171345333      0
#> 182  -0.157192763      0
#> 183  -1.237168132      0
#> 184  -0.754648511      0
#> 185   0.082412912      0
#> 186  -0.222122225      0
#> 187  -2.206044879      0
#> 188  -0.721631370      0
#> 189  -0.027972843      0
#> 190  -0.922742759      0
#> 191  -0.077312228      0
#> 192   0.323136740      0
#> 193  -0.575348299      0
#> 194  -0.740634585      0
#> 195   0.497031583      0
#> 196   0.130689037      0
#> 197   0.492557247      0
#> 198  -0.620894418      0
#> 199  -0.067490363      0
#> 200  -0.950283821      0
#> 201  -1.499206225      0
#> 202  -0.204688899      0
#> 203   0.307852636      0
#> 204  -1.501171580      0
#> 205  -0.158497141      0
#> 206   0.609776620      0
#> 207   0.591830182      0
#> 208  -0.326770178      0
#> 209  -0.706105648      0
#> 210   0.309442887      0
#> 211  -1.398980372      0
#> 212   0.139604131      0
#> 213  -0.188015391      0
#> 214  -0.934642948      0
#> 215  -0.211557991      0
#> 216   1.055899499      0
#> 217   0.307965743      0
#> 218   0.117316992      0
#> 219  -0.985320038      0
#> 220  -0.560435493      0
#> 221   0.523700764      0
#> 222  -0.302368414      0
#> 223   0.959091516      0
#> 224   0.188527471      0
#> 225  -1.279346536      0
#> 226  -0.872438421      0
#> 227  -0.613401970      0
#> 228   0.377915105      0
#> 229  -0.257390382      0
#> 230  -0.337357355      0
#> 231   0.799521634      0
#> 232  -0.334300130      0
#> 233  -1.427939768      0
#> 234   0.528476610      0
#> 235  -0.237279480      0
#> 236  -1.122366424      0
#> 237  -0.478515411      0
#> 238   0.952670656      0
#> 239   0.118767826      0
#> 240   1.080981974      0
#> 241   0.867375559      0
#> 242  -0.412306697      0
#> 243   1.497641813      0
#> 244  -1.428827570      0
#> 245   0.283557193      0
#> 246   0.622919733      0
#> 247   0.867959600      0
#> 248  -0.423832621      0
#> 249  -0.343837867      0
#> 250  -0.172173427      0
#> 251  -0.337851419      0
#> 252  -0.305457093      0
#> 253   0.489521561      0
#> 254   0.034176458      0
#> 255  -0.002603984      0
#> 256  -0.042144018      0
#> 257   0.979674354      0
#> 258   0.894019528      0
#> 259   0.938883648      0
#> 260   1.165835695      0
#> 261   0.264226524      0
#> 262  -0.392450037      0
#> 263   2.523325854      0
#> 264   0.122988814      0
#> 265  -0.636324435      0
#> 266  -1.094031070      0
#> 267  -1.673043916      0
#> 268   0.533036122      0
#> 269   1.274433184      0
#> 270  -0.255492258      0
#> 271  -0.428874444      0
#> 272  -0.009206211      0
#> 273   0.689135674      0
#> 274   0.707585486      0
#> 275   1.186557559      0
#> 276  -0.831601597      0
#> 277  -0.061411754      0
#> 278  -1.067279404      0
#> 279  -1.406893825      0
#> 280  -1.268985714      0
#> 281   0.711023842      0
#> 282  -0.301688859      0
#> 283   1.673320608      0
#> 284   1.584185738      0
#> 285  -1.260647248      0
#> 286   0.648616704      0
#> 287   1.124033193      0
#> 288   0.812544903      0
#> 289   1.038526251      0
#> 290  -1.755430011      0
#> 291   0.490246240      0
#> 292   0.977180285      0
#> 293   1.211547829      0
#> 294   0.307402124      0
#> 295   2.060912115      0
#> 296  -1.073995059      0
#> 297   1.160365017      0
#> 298   1.357104788      0
#> 299  -0.468139704      0
#> 300   0.280024641      0
#> 301  -0.650981918      0
#> 302   2.236355424      0
#> 303  -0.348934985      0
#> 304  -1.316557731      0
#> 305  -1.214699942      0
#> 306  -0.380085855      0
#> 307  -0.247249873      0
#> 308  -0.498059324      0
#> 309  -0.657657697      0
#> 310   0.591668230      0
#> 311  -0.656828842      0
#> 312   0.688044169      0
#> 313  -0.869207096      0
#> 314  -1.101289753      0
#> 315  -1.314352502      0
#> 316  -0.059369938      0
#> 317  -0.665576069      0
#> 318  -1.368742491      0
#> 319  -0.937392271      0
#> 320   0.882253800      0
#> 321   1.417418325      0
#> 322  -0.866122895      0
#> 323  -1.378159710      0
#> 324   0.581140379      0
#> 325   0.947005771      0
#> 326  -1.268657003      0
#> 327   1.029416873      0
#> 328  -0.771894656      0
#> 329  -1.532196650      0
#> 330  -0.402122195      0
#> 331  -1.304352179      0
#> 332   1.992595879      0
#> 333   0.014901305      0
#> 334   0.668607831      0
#> 335   0.743134593      0
#> 336  -0.336696564      0
#> 337   0.771711985      0
#> 338  -0.396733172      0
#> 339  -0.257087281      0
#> 340   0.517430020      0
#> 341   0.150617486      0
#> 342   0.147707098      0
#> 343  -0.609428912      0
#> 344  -0.244623374      0
#> 345   0.304729386      0
#> 346   0.131628199      0
#> 347  -0.968796091      0
#> 348  -0.059787390      0
#> 349  -0.728142769      0
#> 350  -0.517740194      0
#> 351  -2.115829737      0
#> 352  -0.864522848      0
#> 353  -1.788805374      0
#> 354   0.026618857      0
#> 355   1.087199762      0
#> 356   0.169594274      0
#> 357   0.937570768      0
#> 358   0.676344872      0
#> 359  -1.055804132      0
#> 360  -0.384517912      0
#> 361  -1.158367944      0
#> 362  -0.707765621      0
#> 363   0.354170181      0
#> 364   0.299277533      0
#> 365  -2.637681989      0
#> 366  -1.216599232      0
#> 367   2.302170355      0
#> 368   0.549912989      0
#> 369   1.073592067      0
#> 370  -0.791317350      0
#> 371   0.368135439      0
#> 372  -0.417006481      0
#> 373  -1.358009703      0
#> 374  -1.245428047      0
#> 375   0.850037860      0
#> 376  -0.094664503      0
#> 377   2.166728959      0
#> 378  -1.203388803      0
#> 379   1.550842019      0
#> 380   1.445793181      0
#> 381   0.416440961      0
#> 382  -0.144752025      0
#> 383  -1.424181811      0
#> 384  -0.728105535      0
#> 385   0.818597135      0
#> 386   0.333244879      0
#> 387   0.452849680      0
#> 388   0.951337778      0
#> 389   1.129372730      0
#> 390  -0.164447827      0
#> 391   1.225154424      0
#> 392  -0.325240795      0
#> 393   1.301041882      0
#> 394   0.126289384      0
#> 395  -0.155412811      0
#> 396  -0.553024018      0
#> 397  -0.835932742      0
#> 398  -0.844308357      0
#> 399   4.226214554      0
#> 400   0.856443043      0
#> 401   2.367675497      0
#> 402   1.090279059      0
#> 403  -0.174420501      0
#> 404  -0.592311392      0
#> 405  -0.191838160      0
#> 406   0.289372143      0
#> 407  -0.921158436      0
#> 408  -0.316810164      0
#> 409  -0.212657390      0
#> 410  -1.181784575      0
#> 411   0.136465670      0
#> 412  -1.055531661      0
#> 413   0.158981515      0
#> 414   1.168416762      0
#> 415   0.140037560      0
#> 416   0.858186647      0
#> 417   1.911270925      0
#> 418  -1.269484993      0
#> 419  -0.321146772      0
#> 420  -0.574757722      0
#> 421  -0.751758701      0
#> 422   1.034391426      0
#> 423   1.589303982      0
#> 424  -0.138033448      0
#> 425   0.948905113      0
#> 426   0.553531845      0
#> 427   0.477131829      0
#> 428   0.660388389      0
#> 429  -0.135850208      0
#> 430  -2.116056661      0
#> 431   0.788818513      0
#> 432   0.667125155      0
#> 433  -0.092355551      0
#> 434  -0.038169690      0
#> 435   0.037607893      0
#> 436  -0.525994372      0
#> 437  -1.627998655      0
#> 438   0.150231510      0
#> 439   1.951020053      0
#> 440   0.124461931      0
#> 441  -0.503636941      0
#> 442   0.410480200      0
#> 443  -0.210007282      0
#> 444  -0.541250825      0
#> 445   0.120101381      0
#> 446   1.788846529      0
#> 447   0.574489167      0
#> 448  -0.586095258      0
#> 449   1.131781770      0
#> 450   0.903997517      0
#> 451   0.194912913      0
#> 452  -0.584424193      0
#> 453   0.305348721      0
#> 454   0.173871858      0
#> 455   1.556129277      0
#> 456   1.220749093      0
#> 457   0.775138529      0
#> 458  -0.620520077      0
#> 459   1.730653778      0
#> 460   1.467138192      0
#> 461   0.724384947      0
#> 462  -0.146319971      0
#> 463  -1.162300752      0
#> 464  -0.066186256      0
#> 465  -0.550364616      0
#> 466  -0.822005881      0
#> 467   0.421253625      0
#> 468   0.266136804      0
#> 469  -0.138645007      0
#> 470   2.219162919      0
#> 471  -1.628973243      0
#> 472  -2.347974539      0
#> 473   1.509828260      0
#> 474   0.188254634      0
#> 475   0.728876695      0
#> 476   0.082512760      0
#> 477   1.239159243      0
#> 478   0.100548468      0
#> 479   1.111262877      0
#> 480   0.957119659      0
#> 481  -0.434718490      0
#> 482   0.663476773      0
#> 483   0.481369425      0
#> 484  -0.837470787      0
#> 485  -0.066945034      0
#> 486  -0.823064014      0
#> 487   0.284462594      0
#> 488   0.775240066      0
#> 489   0.480872775      0
#> 490   0.880146950      0
#> 491  -0.501462135      0
#> 492  -0.319993250      0
#> 493   0.718975728      0
#> 494   0.733724040      0
#> 495   0.885030029      0
#> 496   1.192654566      0
#> 497   1.449070201      0
#> 498   2.315003379      0
#> 499  -0.050722452      0
#> 500   0.637919516      0
#> 501  -0.284482142      0
#> 502  -0.833469148      0
#> 503  -0.977240302      0
#> 504   0.459334160      0
#> 505  -0.443243558      0
#> 506  -1.449602031      0
#> 507  -1.020031107      0
#> 508  -0.348013680      0
#> 509   1.035149702      0
#> 510  -0.016990652      0
#> 511   0.114274999      0
#> 512  -0.970856833      0
#> 513  -1.248226823      0
#> 514  -0.491782970      0
#> 515  -0.425214923      0
#> 516   1.221825338      0
#> 517   0.860083256      0
#> 518   0.005247856      0
#> 519   0.174910997      0
#> 520   1.320532947      0
#> 521  -1.300567282      0
#> 522  -0.535018563      0
#> 523   0.059631480      0
#> 524  -0.086730157      0
#> 525  -0.034623073      0
#> 526  -1.679770333      0
#> 527  -1.519105008      0
#> 528  -1.930815721      0
#> 529  -1.279323967      0
#> 530   0.066142935      0
#> 531  -0.658490040      0
#> 532  -2.506922529      0
#> 533  -1.501347909      0
#> 534  -0.012279182      0
#> 535  -0.148603154      0
#> 536  -0.226733894      0
#> 537   2.306991673      0
#> 538  -1.580969987      0
#> 539   1.401345196      0
#> 540  -0.056301913      0
#> 541   0.260169941      0
#> 542   0.234463118      0
#> 543   0.550479904      0
#> 544   0.918602855      0
#> 545  -0.812191487      0
#> 546   2.197412378      0
#> 547   0.058875511      0
#> 548  -0.922593824      0
#> 549  -0.376611132      0
#> 550  -1.963130751      0
#> 551   1.689281624      0
#> 552  -0.817430428      0
#> 553   0.015709627      0
#> 554  -0.325364269      0
#> 555   0.661565953      0
#> 556   0.022709447      0
#> 557   0.248319114      0
#> 558   0.079480517      0
#> 559  -1.818443971      0
#> 560  -0.533327516      0
#> 561   0.785376417      0
#> 562   0.608505411      0
#> 563  -1.967733514      0
#> 564   0.221436755      0
#> 565   0.389566550      0
#> 566   1.790869654      0
#> 567  -0.980737824      0
#> 568  -0.161335876      0
#> 569  -0.752435494      0
#> 570   0.468078400      0
#> 571   1.530776716      0
#> 572  -1.086847443      0
#> 573   1.433355891      0
#> 574   0.608050599      0
#> 575   1.514984746      0
#> 576  -1.389330143      0
#> 577  -0.147275177      0
#> 578   0.294329416      0
#> 579  -0.744962973      0
#> 580  -0.355578465      0
#> 581   0.289591232      0
#> 582   0.729276475      0
#> 583   1.656026997      0
#> 584  -1.890649751      0
#> 585  -0.209018290      0
#> 586   0.576081763      0
#> 587   0.644625290      0
#> 588   0.396310196      0
#> 589   0.432289207      0
#> 590   0.464988935      0
#> 591  -0.163619090      0
#> 592  -0.964610909      0
#> 593  -0.778788731      0
#> 594  -0.195960681      0
#> 595   0.195945819      0
#> 596   1.042879439      0
#> 597  -0.660052815      0
#> 598   0.180083017      0
#> 599  -1.080908514      0
#> 600   1.011370640      0
#> 601  -0.556621342      0
#> 602  -0.369751138      0
#> 603   0.729904874      0
#> 604   0.627835820      0
#> 605  -0.630111413      0
#> 606  -0.964623433      0
#> 607   0.864921929      0
#> 608   0.596074511      0
#> 609   0.472776042      0
#> 610   1.541432650      0
#> 611  -1.714831608      0
#> 612   0.024040133      0
#> 613  -1.004255470      0
#> 614  -1.371700414      0
#> 615   1.157945869      0
#> 616  -0.125240223      0
#> 617  -0.305058757      0
#> 618  -0.378946547      0
#> 619   0.451782415      0
#> 620  -1.359440438      0
#> 621   0.788041316      0
#> 622  -1.001746601      0
#> 623   0.614825464      0
#> 624   2.770730580      0
#> 625   1.446095283      0
#> 626  -0.338533820      0
#> 627  -0.893152788      0
#> 628  -1.881409954      0
#> 629  -1.011273477      0
#> 630   0.062169923      0
#> 631   0.734886203      0
#> 632  -0.750140838      0
#> 633  -0.749214707      0
#> 634  -0.143062381      0
#> 635  -0.399132250      0
#> 636   0.306788730      0
#> 637   1.240202653      0
#> 638   0.779413733      0
#> 639   1.422740708      0
#> 640   0.719474646      0
#> 641  -0.503707577      0
#> 642  -0.520213040      0
#> 643   1.190080141      0
#> 644   0.479625923      0
#> 645   0.480521162      0
#> 646  -0.260394320      0
#> 647  -0.815584506      0
#> 648  -0.242960295      0
#> 649   1.639059784      0
#> 650   1.462246986      0
#> 651  -0.980629360      0
#> 652   1.615532011      0
#> 653   1.055534802      0
#> 654  -0.482238479      0
#> 655  -0.824800075      0
#> 656  -1.109816984      0
#> 657   0.085685339      0
#> 658   0.477549116      0
#> 659   1.060583820      0
#> 660   0.582651155      0
#> 661  -1.023986457      0
#> 662  -0.308518708      0
#> 663  -1.158559968      0
#> 664  -0.749826199      0
#> 665   0.003057177      0
#> 666   0.793030121      0
#> 667   0.016074340      0
#> 668  -0.561991710      0
#> 669  -0.399509748      0
#> 670  -0.436852127      0
#> 671  -3.023544459      0
#> 672  -1.228058050      0
#> 673  -1.234633067      0
#> 674   0.806257163      0
#> 675   1.311991576      0
#> 676  -1.666030380      0
#> 677  -0.219900404      0
#> 678  -1.064425592      0
#> 679  -0.532292942      0
#> 680   0.431350216      0
#> 681   0.013224551      0
#> 682  -2.248206304      0
#> 683  -0.101585871      0
#> 684  -0.886710572      0
#> 685  -0.094688006      0
#> 686  -0.513372045      0
#> 687  -0.847814162      0
#> 688   0.809479755      0
#> 689  -2.360272347      0
#> 690   0.235037806      0
#> 691  -0.678842329      0
#> 692   0.588122287      0
#> 693  -0.359681503      0
#> 694  -1.194234236      0
#> 695   0.397154293      0
#> 696   0.250192868      0
#> 697  -1.078538202      0
#> 698  -0.938347153      0
#> 699   0.015127511      0
#> 700   1.571243664      0
#> 701   0.666416046      0
#> 702   0.138267148      0
#> 703   0.757283499      0
#> 704  -0.751005808      0
#> 705   0.240952822      0
#> 706  -0.493731688      0
#> 707   1.789805805      0
#> 708  -0.694375231      0
#> 709  -0.196048286      0
#> 710   0.515045066      0
#> 711  -2.412391366      0
#> 712  -0.136844027      0
#> 713   0.747096891      0
#> 714   1.329629792      0
#> 715   0.244571990      0
#> 716  -1.944975700      0
#> 717   0.476742709      0
#> 718   1.672209858      0
#> 719   3.392894711      0
#> 720  -0.236903882      0
#> 721  -0.545084117      0
#> 722   0.759189508      0
#> 723   1.126418854      0
#> 724   0.021558281      0
#> 725  -0.427274104      0
#> 726  -1.795872868      0
#> 727   0.590457874      0
#> 728   2.514422021      0
#> 729   0.471467850      0
#> 730   0.336188595      0
#> 731   0.289864358      0
#> 732   1.014361252      0
#> 733   0.948562629      0
#> 734   1.429773744      0
#> 735   0.016284859      0
#> 736   0.977355309      0
#> 737  -1.331820349      0
#> 738   0.431318967      0
#> 739  -0.323246387      0
#> 740   0.477652304      0
#> 741  -2.207215002      0
#> 742  -0.589239885      0
#> 743  -0.621516488      0
#> 744   0.774678342      0
#> 745  -1.649856014      0
#> 746   1.055259042      0
#> 747   0.726004713      0
#> 748  -0.335656258      0
#> 749   0.916629271      0
#> 750  -0.206885169      0
#> 751  -2.056903618      0
#> 752  -0.777111609      0
#> 753  -1.403971881      0
#> 754  -0.329481029      0
#> 755  -1.414925653      0
#> 756   1.004619462      0
#> 757  -0.183341459      0
#> 758  -0.739368178      0
#> 759  -0.235284926      0
#> 760   0.084783096      0
#> 761   0.344000522      0
#> 762   1.333282573      0
#> 763  -0.713650862      0
#> 764   0.464861467      0
#> 765   0.242056463      0
#> 766  -0.197548479      0
#> 767  -0.513310187      0
#> 768  -0.053545594      0
#> 769  -0.668738267      0
#> 770   1.236268474      0
#> 771   0.154489669      0
#> 772   0.179872702      0
#> 773   0.669226413      0
#> 774   0.992818199      0
#> 775  -0.256036725      0
#> 776   1.253870259      0
#> 777   1.625792900      0
#> 778  -0.367854340      0
#> 779   1.419320904      0
#> 780  -0.635847699      0
#> 781  -0.106852122      0
#> 782   0.779149486      0
#> 783   1.100590979      0
#> 784  -0.317926291      0
#> 785   0.674209238      0
#> 786   0.103494779      0
#> 787   1.109531907      0
#> 788  -0.317340083      0
#> 789   1.868293901      0
#> 790   0.699209101      0
#> 791   0.032016993      0
#> 792  -0.374918716      0
#> 793   0.585089249      0
#> 794   0.923634827      0
#> 795   0.767009154      0
#> 796   1.203615373      0
#> 797  -0.189584230      0
#> 798   0.188671296      0
#> 799  -1.138609298      0
#> 800   0.823715620      0
#> 801  -1.157239369      0
#> 802   0.558369979      0
#> 803   0.668126718      0
#> 804   0.130921847      0
#> 805  -0.393412645      0
#> 806  -1.586576797      0
#> 807  -0.975928757      0
#> 808   1.274820014      0
#> 809  -0.286047065      0
#> 810   0.262618712      0
#> 811   0.507576758      0
#> 812   1.384461598      0
#> 813   0.752346399      0
#> 814  -0.698254052      0
#> 815   0.292554337      0
#> 816  -1.009368871      0
#> 817  -1.342646256      0
#> 818   0.395283783      0
#> 819   1.307543944      0
#> 820   1.013531687      0
#> 821   0.833569146      0
#> 822   2.109857444      0
#> 823  -0.894334583      0
#> 824   0.779678282      0
#> 825   0.025268930      0
#> 826   0.615594786      0
#> 827  -0.157248500      0
#> 828  -0.036393203      0
#> 829  -0.509281912      0
#> 830   0.818828372      0
#> 831  -1.090789358      0
#> 832  -0.540453008      0
#> 833   1.053092338      0
#> 834  -0.629852796      0
#> 835  -0.770188483      0
#> 836   1.653466959      0
#> 837  -0.473253010      0
#> 838   0.098710636      0
#> 839  -0.083764829      0
#> 840   1.410980857      0
#> 841   0.707527566      0
#> 842   0.703971188      0
#> 843   0.733913235      0
#> 844  -1.615585317      0
#> 845   1.071228474      0
#> 846  -1.966245592      0
#> 847  -0.292885535      0
#> 848  -0.752537370      0
#> 849   0.091331261      0
#> 850   1.577868618      0
#> 851  -0.550785258      0
#> 852   0.080464647      0
#> 853   1.359005516      0
#> 854  -0.927836662      0
#> 855   0.674288420      0
#> 856   0.267430987      0
#> 857  -0.950898594      0
#> 858   0.800198259      0
#> 859   0.326834305      0
#> 860  -1.153587462      0
#> 861   0.853502220      0
#> 862  -0.864036281      0
#> 863  -2.081407551      0
#> 864   1.102869954      0
#> 865   0.679525003      0
#> 866  -0.038018424      0
#> 867   1.922723664      0
#> 868   0.246603703      0
#> 869  -1.132854527      0
#> 870  -0.122388781      0
#> 871  -0.335395321      0
#> 872   0.630821360      0
#> 873   1.000590112      0
#> 874  -0.940128974      0
#> 875   1.277946188      0
#> 876  -1.450015306      0
#> 877   1.168138220      0
#> 878   0.396764875      0
#> 879   0.472686366      0
#> 880  -0.745875793      0
#> 881  -1.041737994      0
#> 882  -0.159106109      0
#> 883   1.486201715      0
#> 884  -0.172771121      0
#> 885   1.458309993      0
#> 886   0.480125017      0
#> 887  -0.449278791      0
#> 888  -1.416761398      0
#> 889  -1.059504397      0
#> 890  -0.271121908      0
#> 891  -1.090587908      0
#> 892   1.153988863      0
#> 893  -0.524905689      0
#> 894   1.811728024      0
#> 895   0.216632867      0
#> 896   1.970759310      0
#> 897  -0.356427975      0
#> 898  -0.448708762      0
#> 899   1.146563454      0
#> 900   1.254638083      0
#> 901   1.411693726      1
#> 902   0.039312303      1
#> 903  -0.079185719      1
#> 904   2.637045887      1
#> 905   1.056034402      1
#> 906   1.265937576      1
#> 907   2.473230140      1
#> 908   2.139228406      1
#> 909   1.801338371      1
#> 910   2.971259780      1
#> 911   1.428106719      1
#> 912   0.554308905      1
#> 913   2.524212064      1
#> 914   1.187990250      1
#> 915   0.777854901      1
#> 916   0.607550418      1
#> 917  -0.585510756      1
#> 918   0.067303319      1
#> 919   2.044904668      1
#> 920  -0.327493016      1
#> 921  -0.010173161      1
#> 922   2.251197504      1
#> 923   2.080564235      1
#> 924   1.760116274      1
#> 925   2.224645117      1
#> 926   0.096020586      1
#> 927   1.280218696      1
#> 928   1.862314284      1
#> 929   1.884087329      1
#> 930   0.382805512      1
#> 931  -0.470883817      1
#> 932   0.617623156      1
#> 933   2.287539799      1
#> 934   0.401818009      1
#> 935   0.590954796      1
#> 936   1.617640851      1
#> 937   1.682303656      1
#> 938  -0.469022027      1
#> 939   0.752187588      1
#> 940   0.271767864      1
#> 941   1.046225803      1
#> 942   2.746504484      1
#> 943   1.208932723      1
#> 944   1.790046868      1
#> 945   2.008003109      1
#> 946   1.099820434      1
#> 947   2.185655729      1
#> 948   2.668251558      1
#> 949  -0.114138822      1
#> 950   0.898553578      1
#> 951   1.467509675      1
#> 952   1.702065263      1
#> 953   0.537709560      1
#> 954   2.034058874      1
#> 955   1.556764592      1
#> 956   0.712180691      1
#> 957   0.949505736      1
#> 958   0.333876522      1
#> 959   0.376207702      1
#> 960   1.317086951      1
#> 961  -0.446466223      1
#> 962   0.049898401      1
#> 963  -0.566145271      1
#> 964   1.161487925      1
#> 965  -1.014965101      1
#> 966  -0.266233528      1
#> 967  -0.357069234      1
#> 968   0.515256415      1
#> 969   0.787618259      1
#> 970   3.012297360      1
#> 971   0.356938163      1
#> 972   2.528241856      1
#> 973   0.727323315      1
#> 974   1.064449698      1
#> 975   0.682487912      1
#> 976   2.454978632      1
#> 977  -0.867181324      1
#> 978   0.978186040      1
#> 979   0.432663748      1
#> 980   2.293084469      1
#> 981   0.470324551      1
#> 982   0.539792240      1
#> 983   0.508365884      1
#> 984   0.048326614      1
#> 985   2.479309255      1
#> 986  -0.382255441      1
#> 987   1.174025237      1
#> 988  -0.036062573      1
#> 989   1.302552939      1
#> 990   0.046774359      1
#> 991   2.120284664      1
#> 992   2.271180809      1
#> 993   0.265803052      1
#> 994   0.170108267      1
#> 995   1.346560127      1
#> 996   0.012196131      1
#> 997   1.087137348      1
#> 998   1.905992744      1
#> 999  -0.187801450      1
#> 1000  0.067347700      1