Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     0.0546301058      0
#> 2     0.6688372201      0
#> 3    -1.0239720727      0
#> 4    -1.4140642317      0
#> 5     0.0097034151      0
#> 6    -0.5469646632      0
#> 7    -0.8789657633      0
#> 8    -1.0957487237      0
#> 9     0.2635576045      0
#> 10    0.6210096324      0
#> 11    1.0488461119      0
#> 12    0.7589277967      0
#> 13    2.6769778065      0
#> 14    1.3478683533      0
#> 15   -0.0471938544      0
#> 16    1.8305157664      0
#> 17   -1.4296248841      0
#> 18   -1.2616685572      0
#> 19   -0.2896499062      0
#> 20    0.0448865427      0
#> 21   -0.5939367581      0
#> 22   -1.0042695066      0
#> 23    0.0194696167      0
#> 24   -0.0881104663      0
#> 25    0.5034983259      0
#> 26    0.5144614433      0
#> 27    0.5090073712      0
#> 28    0.6127735843      0
#> 29    0.6693684143      0
#> 30    1.1871523384      0
#> 31   -0.2874918102      0
#> 32    0.2724552054      0
#> 33   -0.2712638700      0
#> 34   -0.9443511384      0
#> 35    0.7861281269      0
#> 36   -0.9228725975      0
#> 37   -1.7315532953      0
#> 38   -0.3256211814      0
#> 39   -0.2539675920      0
#> 40   -0.8463438757      0
#> 41    0.1552732753      0
#> 42   -0.7957293950      0
#> 43    1.0212925465      0
#> 44   -1.0283770012      0
#> 45   -0.4449007684      0
#> 46   -0.7648754695      0
#> 47    0.1113152435      0
#> 48    0.4591886358      0
#> 49    0.9345673759      0
#> 50    0.9292091534      0
#> 51   -0.0816263855      0
#> 52   -0.8655474438      0
#> 53   -0.6767253484      0
#> 54   -0.0654206140      0
#> 55   -0.0867340930      0
#> 56   -2.4331573924      0
#> 57   -0.6609672227      0
#> 58    0.7256998354      0
#> 59    0.0869630718      0
#> 60    0.1162865837      0
#> 61    0.9938599518      0
#> 62   -1.3069272545      0
#> 63    1.4208920631      0
#> 64   -0.3721677654      0
#> 65   -0.2466925423      0
#> 66   -0.8542700851      0
#> 67    0.7755448031      0
#> 68    0.6390486537      0
#> 69    0.2770128640      0
#> 70   -0.5978834899      0
#> 71    0.8868952477      0
#> 72   -0.5056251051      0
#> 73    0.9260455291      0
#> 74   -1.0094548065      0
#> 75    1.7813246169      0
#> 76   -0.8753823993      0
#> 77   -0.7302059775      0
#> 78    0.1631639178      0
#> 79    1.5961059291      0
#> 80   -0.2707937076      0
#> 81   -1.5767844429      0
#> 82   -0.1348688092      0
#> 83   -0.0294804267      0
#> 84   -1.6043716161      0
#> 85   -1.3208394915      0
#> 86    1.6608504620      0
#> 87   -1.4640613686      0
#> 88    0.3286001136      0
#> 89   -1.2603233351      0
#> 90   -0.3261714029      0
#> 91   -0.5416715712      0
#> 92    1.7033759810      0
#> 93    1.6958949586      0
#> 94    1.4004577572      0
#> 95   -0.6908663936      0
#> 96    0.9153363880      0
#> 97   -0.5947201596      0
#> 98    1.0874822361      0
#> 99   -0.6855105990      0
#> 100  -0.0355763896      0
#> 101  -1.1600868134      0
#> 102  -0.1387582969      0
#> 103   0.2268888540      0
#> 104  -0.6345794075      0
#> 105  -0.0256906881      0
#> 106   0.1025458245      0
#> 107  -0.2593374396      0
#> 108  -1.5909945221      0
#> 109  -0.5751213268      0
#> 110  -1.1428244292      0
#> 111  -1.1297295367      0
#> 112  -1.3356425559      0
#> 113  -0.8133737117      0
#> 114  -0.1853150292      0
#> 115   0.7066078618      0
#> 116  -0.7933977301      0
#> 117   1.1920883006      0
#> 118  -0.8017014425      0
#> 119  -0.2475886785      0
#> 120   0.3354479540      0
#> 121   1.1454390237      0
#> 122   0.3618360024      0
#> 123   1.0480452470      0
#> 124  -1.3380520771      0
#> 125   0.1576422221      0
#> 126  -0.6540079308      0
#> 127   0.1720674180      0
#> 128   0.3631872012      0
#> 129  -0.0515865423      0
#> 130   0.9956738526      0
#> 131  -1.6323603081      0
#> 132  -0.2863646609      0
#> 133   0.6021480496      0
#> 134  -1.0546517562      0
#> 135   0.6712538993      0
#> 136   0.0491486667      0
#> 137  -0.3278299354      0
#> 138   0.7327904761      0
#> 139  -1.0574138081      0
#> 140   0.3364838122      0
#> 141   0.1372809835      0
#> 142   1.4200977999      0
#> 143   1.0496758411      0
#> 144  -1.6925676560      0
#> 145  -0.0140372644      0
#> 146   0.3225755294      0
#> 147  -0.9223696795      0
#> 148  -0.9427963019      0
#> 149  -1.0902348903      0
#> 150  -1.9516106257      0
#> 151   2.0378367787      0
#> 152   1.4930934406      0
#> 153   0.6394302927      0
#> 154  -1.2618158654      0
#> 155  -0.7585835216      0
#> 156  -0.3476510845      0
#> 157  -1.4303989007      0
#> 158  -1.5329994790      0
#> 159   0.1432128964      0
#> 160  -1.3676537644      0
#> 161  -1.1918025635      0
#> 162   0.3877352660      0
#> 163  -0.4948116031      0
#> 164   0.0197467286      0
#> 165  -0.0816458207      0
#> 166  -1.1681269404      0
#> 167   2.2938286017      0
#> 168  -1.8261530434      0
#> 169  -1.3457908085      0
#> 170  -0.1394236909      0
#> 171   0.5348129758      0
#> 172  -1.3989686900      0
#> 173   1.0925964420      0
#> 174  -0.0293638171      0
#> 175  -0.2878980765      0
#> 176   0.0151278763      0
#> 177   1.3809444231      0
#> 178  -0.0164313479      0
#> 179  -0.5508234638      0
#> 180  -1.2676938243      0
#> 181   0.8420153379      0
#> 182  -0.4272202937      0
#> 183   1.3987345717      0
#> 184   0.9413754790      0
#> 185  -0.4384159894      0
#> 186   0.8441353172      0
#> 187   0.3245555891      0
#> 188   0.3067296818      0
#> 189  -0.8094894960      0
#> 190   0.2399287835      0
#> 191   0.7655300268      0
#> 192  -0.1576601338      0
#> 193   0.4281820069      0
#> 194  -0.5186470357      0
#> 195   0.0040346123      0
#> 196  -0.3393828669      0
#> 197  -0.5685037716      0
#> 198  -0.5015268753      0
#> 199  -2.1146079926      0
#> 200   0.1272344974      0
#> 201  -1.0135617577      0
#> 202  -0.5596877581      0
#> 203  -0.3342744549      0
#> 204   0.2043391574      0
#> 205   2.0954303142      0
#> 206  -0.4210447204      0
#> 207  -1.3507954648      0
#> 208   1.4435730567      0
#> 209  -0.3123624748      0
#> 210  -0.3051690516      0
#> 211   2.3638320452      0
#> 212  -0.9869748814      0
#> 213  -0.6352298923      0
#> 214   1.7955669233      0
#> 215  -2.2377317547      0
#> 216  -0.3422719268      0
#> 217   0.1968142861      0
#> 218  -0.9080664683      0
#> 219   1.8664167971      0
#> 220  -1.5457897681      0
#> 221   0.3638570388      0
#> 222   0.0841758738      0
#> 223  -0.9506731950      0
#> 224  -1.4596305118      0
#> 225   1.2486231363      0
#> 226  -1.4058472812      0
#> 227  -0.6637849325      0
#> 228  -0.1265170477      0
#> 229  -0.5120503755      0
#> 230   2.7866011206      0
#> 231   2.3258042841      0
#> 232  -0.3412640234      0
#> 233  -0.1608920660      0
#> 234  -0.1242868839      0
#> 235   1.4708039163      0
#> 236  -0.8932651865      0
#> 237  -0.5235975248      0
#> 238  -0.3800777245      0
#> 239  -0.8102350120      0
#> 240   0.4975179568      0
#> 241  -1.9494345730      0
#> 242  -0.7016256789      0
#> 243  -1.8779993127      0
#> 244  -1.0433231571      0
#> 245   0.4840598878      0
#> 246   0.4307095385      0
#> 247  -0.3982911455      0
#> 248  -0.5943005878      0
#> 249  -0.5179885002      0
#> 250  -1.0509178316      0
#> 251  -0.7625929582      0
#> 252   1.0820262330      0
#> 253  -1.6306105610      0
#> 254   2.2282418517      0
#> 255  -0.0872125001      0
#> 256   1.7201950084      0
#> 257   1.5235784902      0
#> 258  -1.4313726623      0
#> 259  -0.9446952855      0
#> 260  -0.3182673293      0
#> 261   0.1397802495      0
#> 262   0.6377908835      0
#> 263   1.2581311778      0
#> 264  -0.5625764792      0
#> 265  -0.4349640808      0
#> 266   1.2695567720      0
#> 267  -0.8812886661      0
#> 268  -0.5424288028      0
#> 269   1.9267157154      0
#> 270   2.2771543523      0
#> 271   0.1918394028      0
#> 272  -0.3559403200      0
#> 273  -1.0149456893      0
#> 274  -0.0043296733      0
#> 275   0.0924940180      0
#> 276  -0.6581525285      0
#> 277  -0.7032172669      0
#> 278   0.8567129851      0
#> 279   1.0026098027      0
#> 280   1.2668584961      0
#> 281   1.1085321281      0
#> 282  -1.9304400392      0
#> 283   0.7833405706      0
#> 284   0.0940410612      0
#> 285  -0.3099864951      0
#> 286   1.1109583147      0
#> 287  -1.2147010931      0
#> 288   0.3693307419      0
#> 289  -0.7784967163      0
#> 290   0.8797495399      0
#> 291   1.0613527140      0
#> 292  -0.1059452661      0
#> 293   0.3314393574      0
#> 294   1.1161895354      0
#> 295  -0.1900986077      0
#> 296   0.8583841996      0
#> 297   0.3931759885      0
#> 298  -1.1031268043      0
#> 299  -0.1981526499      0
#> 300  -0.6372159827      0
#> 301   0.8842498983      0
#> 302  -0.3641680151      0
#> 303   1.5109762104      0
#> 304   0.5145101902      0
#> 305  -0.6752339622      0
#> 306   0.0742049039      0
#> 307  -1.5559984056      0
#> 308  -0.9303354167      0
#> 309   0.8845995664      0
#> 310  -1.1328374417      0
#> 311   1.4110845938      0
#> 312   0.3347017503      0
#> 313   0.7576256726      0
#> 314  -0.6157192686      0
#> 315  -0.6322733726      0
#> 316  -0.4914200310      0
#> 317  -1.3275630937      0
#> 318   0.0126233747      0
#> 319   0.4660338301      0
#> 320  -0.6676158622      0
#> 321   0.7705273850      0
#> 322  -0.1057503688      0
#> 323  -0.0473244557      0
#> 324   1.4278757813      0
#> 325  -0.5276693998      0
#> 326   0.3576690625      0
#> 327  -0.4232639885      0
#> 328   1.9301039673      0
#> 329  -0.4094390296      0
#> 330   3.1994670743      0
#> 331  -1.7542465543      0
#> 332  -0.3354270817      0
#> 333   0.2498731165      0
#> 334   0.5294204733      0
#> 335  -1.3143679574      0
#> 336   0.6750747132      0
#> 337  -1.7443614552      0
#> 338  -0.3183320318      0
#> 339   0.5049535801      0
#> 340  -0.6923419391      0
#> 341  -0.2813615802      0
#> 342   0.1035851200      0
#> 343   1.4708013900      0
#> 344   0.5645859993      0
#> 345  -0.2211178833      0
#> 346  -0.0384369261      0
#> 347  -0.1513042638      0
#> 348  -0.7763682797      0
#> 349   2.3814785463      0
#> 350  -0.9616768094      0
#> 351   0.2644898654      0
#> 352   0.9884767163      0
#> 353  -2.0963672504      0
#> 354  -0.3732821878      0
#> 355   0.3712996985      0
#> 356   1.3161585404      0
#> 357   0.0522465045      0
#> 358   0.2059817648      0
#> 359   0.4573322728      0
#> 360   0.2258030446      0
#> 361   1.5947370000      0
#> 362  -0.5724593333      0
#> 363   0.3189305976      0
#> 364   1.5539778548      0
#> 365   0.3115312643      0
#> 366   1.2400604842      0
#> 367  -0.7826355499      0
#> 368  -0.7238350785      0
#> 369  -0.5898250088      0
#> 370  -0.6983164271      0
#> 371   0.1535405328      0
#> 372   1.1926075046      0
#> 373  -0.5778617757      0
#> 374   0.0675066532      0
#> 375  -1.3200321380      0
#> 376   2.2269001069      0
#> 377   0.7869581567      0
#> 378  -0.7457835647      0
#> 379  -0.0443439793      0
#> 380  -1.0788301340      0
#> 381  -0.3515010105      0
#> 382  -0.7263667899      0
#> 383   0.7093331332      0
#> 384  -1.6980427467      0
#> 385   0.2753252777      0
#> 386  -0.9195311554      0
#> 387   0.0344625207      0
#> 388  -1.0297419010      0
#> 389   0.7629019968      0
#> 390   0.0965288914      0
#> 391   2.2511857861      0
#> 392   0.9251540622      0
#> 393   0.1098434155      0
#> 394   0.2740538474      0
#> 395  -0.4875285919      0
#> 396   0.3113972521      0
#> 397   0.1333816698      0
#> 398   1.8388026358      0
#> 399  -1.0521496045      0
#> 400  -1.7023137407      0
#> 401   0.7633373763      0
#> 402   0.9723674070      0
#> 403  -0.2765476075      0
#> 404   0.7755620879      0
#> 405  -0.1375976211      0
#> 406   0.5823158607      0
#> 407   1.1266424017      0
#> 408   0.7390822418      0
#> 409   0.4655388579      0
#> 410   0.7571666684      0
#> 411  -0.7599754386      0
#> 412   0.7037882365      0
#> 413  -1.4007733012      0
#> 414  -0.1470741291      0
#> 415   0.9122168077      0
#> 416   1.7978477822      0
#> 417   1.2981480542      0
#> 418   0.2855438054      0
#> 419   0.4325935358      0
#> 420  -0.6653631846      0
#> 421  -1.5315672471      0
#> 422  -0.4731516491      0
#> 423  -1.1968886823      0
#> 424   2.1133538454      0
#> 425  -1.3145618585      0
#> 426  -2.9240282262      0
#> 427   0.1208919889      0
#> 428  -0.7979532962      0
#> 429  -1.7557969151      0
#> 430   0.0500236797      0
#> 431   1.9481443273      0
#> 432   0.6545195774      0
#> 433   0.1397910253      0
#> 434   0.3898159888      0
#> 435   0.7771585995      0
#> 436   0.0552306901      0
#> 437  -0.7123897910      0
#> 438   0.3882555563      0
#> 439  -0.4963457553      0
#> 440  -1.5689358533      0
#> 441  -0.6078184895      0
#> 442  -1.2430893179      0
#> 443   0.5443132969      0
#> 444  -0.5057675791      0
#> 445  -0.3475603783      0
#> 446   0.2278813986      0
#> 447   2.0125150386      0
#> 448   0.1812111882      0
#> 449   0.2366474164      0
#> 450   1.7226948969      0
#> 451  -0.2909398346      0
#> 452   0.4669641347      0
#> 453   0.0977633252      0
#> 454  -0.8582635151      0
#> 455   1.4761785496      0
#> 456   1.1010698563      0
#> 457   0.1170836084      0
#> 458   1.5947497616      0
#> 459  -1.4444821398      0
#> 460  -0.5919458908      0
#> 461   2.0562337462      0
#> 462   0.4387764555      0
#> 463  -0.6468731183      0
#> 464   0.3954236700      0
#> 465   0.3493833429      0
#> 466  -1.2601715390      0
#> 467   1.3655793956      0
#> 468  -0.4892967921      0
#> 469  -0.1944409320      0
#> 470  -0.5453641624      0
#> 471  -0.0449471407      0
#> 472   0.3086107816      0
#> 473   2.2248313196      0
#> 474  -0.2974896467      0
#> 475  -0.8335945533      0
#> 476   0.9674219889      0
#> 477   0.9887799644      0
#> 478  -1.5709824347      0
#> 479   0.0408377132      0
#> 480  -2.7062990842      0
#> 481  -0.2256037261      0
#> 482   0.9264069586      0
#> 483   0.5304336207      0
#> 484   1.1008871445      0
#> 485  -1.2926909800      0
#> 486   0.1090579416      0
#> 487  -0.0371998035      0
#> 488   0.7146150263      0
#> 489  -0.2564479216      0
#> 490   0.8256997531      0
#> 491   0.7434798073      0
#> 492  -0.5735796325      0
#> 493   1.6163311979      0
#> 494  -1.5049074504      0
#> 495   0.1914297534      0
#> 496   0.2541908070      0
#> 497   0.1160420893      0
#> 498  -1.3591401490      0
#> 499   0.3049931632      0
#> 500   0.1555648170      0
#> 501   0.4716713492      0
#> 502  -0.2907281025      0
#> 503   0.6369128570      0
#> 504   0.6087805150      0
#> 505   0.1253554256      0
#> 506  -0.9741220614      0
#> 507   0.1226135904      0
#> 508   0.3317188160      0
#> 509   0.0898152187      0
#> 510  -0.3405246432      0
#> 511  -0.2437883368      0
#> 512  -0.3984719898      0
#> 513  -0.1981105552      0
#> 514  -0.7710652593      0
#> 515   0.5316309584      0
#> 516  -0.0749431369      0
#> 517  -2.4420088842      0
#> 518   0.4044865529      0
#> 519   0.1401330780      0
#> 520  -0.3141916794      0
#> 521   0.0378248450      0
#> 522  -0.4512337537      0
#> 523   1.3421278010      0
#> 524  -0.4412432051      0
#> 525   0.8556572691      0
#> 526   0.9878449113      0
#> 527  -1.0261298521      0
#> 528   0.6086737460      0
#> 529   0.3049507917      0
#> 530   1.6967464376      0
#> 531  -0.7767701297      0
#> 532  -1.1404693243      0
#> 533  -0.5216546910      0
#> 534  -0.9290966515      0
#> 535  -1.1871293144      0
#> 536  -0.3508076971      0
#> 537   0.4355066847      0
#> 538  -0.4068572251      0
#> 539  -0.5171542931      0
#> 540   0.4146597497      0
#> 541   0.0630071761      0
#> 542  -1.4198452652      0
#> 543   1.2926842613      0
#> 544  -1.0736361825      0
#> 545  -1.2434908011      0
#> 546  -0.9821840413      0
#> 547  -0.1443357435      0
#> 548   0.7778720226      0
#> 549  -0.2489571298      0
#> 550   1.4614242382      0
#> 551   0.9011948923      0
#> 552   0.8041382849      0
#> 553  -0.1496414040      0
#> 554   1.5618548101      0
#> 555  -2.4833466605      0
#> 556   2.4179553990      0
#> 557  -1.0193806394      0
#> 558   0.9638150404      0
#> 559   1.0596205792      0
#> 560  -1.5675997717      0
#> 561  -1.2684510149      0
#> 562  -0.0424104965      0
#> 563  -0.2838968722      0
#> 564   0.4007106128      0
#> 565   0.1073402206      0
#> 566  -0.8192445796      0
#> 567  -1.5723915389      0
#> 568  -1.8858969016      0
#> 569  -0.4996413166      0
#> 570  -2.7008477074      0
#> 571  -0.1353665877      0
#> 572   0.9927663331      0
#> 573  -0.7817069278      0
#> 574   0.8150893722      0
#> 575   1.0008206984      0
#> 576   0.6203772999      0
#> 577  -1.5810420149      0
#> 578  -1.2154166469      0
#> 579  -1.1630536095      0
#> 580  -0.0790484715      0
#> 581  -1.1529416497      0
#> 582   1.1561555072      0
#> 583   1.3189977679      0
#> 584  -0.6413527412      0
#> 585  -0.2976263189      0
#> 586  -0.1815806971      0
#> 587   0.8395847735      0
#> 588  -0.1659758594      0
#> 589   0.7636357131      0
#> 590  -1.9109075518      0
#> 591   0.5851349941      0
#> 592  -1.3038966428      0
#> 593  -0.2651218530      0
#> 594  -0.5943490310      0
#> 595  -0.7841153431      0
#> 596  -0.4490557170      0
#> 597  -0.5823821173      0
#> 598   0.7325618851      0
#> 599  -0.9057676597      0
#> 600  -0.2469004842      0
#> 601   1.3830753741      0
#> 602  -1.4786242501      0
#> 603   0.3313443706      0
#> 604   0.9654062663      0
#> 605  -0.5298847796      0
#> 606   0.5344155390      0
#> 607  -1.4479831144      0
#> 608  -0.0878293602      0
#> 609  -1.0577153162      0
#> 610   0.2202160006      0
#> 611   0.8799704952      0
#> 612   0.4339252006      0
#> 613  -0.4542160925      0
#> 614   0.8708092753      0
#> 615   0.1313891025      0
#> 616   1.4725212639      0
#> 617  -0.2471145817      0
#> 618  -0.0835400492      0
#> 619   0.9676670306      0
#> 620  -0.7704163089      0
#> 621   0.5230094432      0
#> 622  -0.1640872077      0
#> 623  -0.6358022886      0
#> 624   0.8849826521      0
#> 625   0.7561028547      0
#> 626  -0.0261310363      0
#> 627   1.2412782860      0
#> 628   0.7521472504      0
#> 629  -0.3342187608      0
#> 630  -1.5218298902      0
#> 631   0.7065051099      0
#> 632  -0.8956588707      0
#> 633   0.9167066196      0
#> 634   0.3894380495      0
#> 635  -1.9104408158      0
#> 636   1.5668386076      0
#> 637  -0.1604586541      0
#> 638   1.1512105435      0
#> 639   0.1506419830      0
#> 640   0.1113493278      0
#> 641  -0.3909387015      0
#> 642   0.2876881422      0
#> 643   0.3003940992      0
#> 644   1.2155672485      0
#> 645  -0.2710705012      0
#> 646  -0.0664774704      0
#> 647   2.0325651781      0
#> 648   0.5139602762      0
#> 649  -1.3771146807      0
#> 650  -1.5522423134      0
#> 651  -0.4424346914      0
#> 652   0.5004041198      0
#> 653  -0.2962644348      0
#> 654  -1.3600267532      0
#> 655  -0.3055047989      0
#> 656   0.3686271355      0
#> 657  -0.0963000832      0
#> 658  -0.1131541133      0
#> 659   1.2719672806      0
#> 660   1.2901697702      0
#> 661   0.6940581139      0
#> 662   0.2630950616      0
#> 663   2.0279390185      0
#> 664   0.1911749877      0
#> 665  -1.8027688197      0
#> 666  -0.6079178157      0
#> 667  -0.9587489549      0
#> 668  -0.6406720026      0
#> 669   0.0995250726      0
#> 670  -0.0598506182      0
#> 671   1.9740936695      0
#> 672  -1.6295080788      0
#> 673  -0.3845233122      0
#> 674   1.0434625297      0
#> 675   1.6169628725      0
#> 676   1.2682999525      0
#> 677  -0.3995730821      0
#> 678   2.7366823751      0
#> 679  -0.0022875303      0
#> 680   0.7923461618      0
#> 681   0.0405039047      0
#> 682  -0.2260665880      0
#> 683  -0.1760545389      0
#> 684  -0.9638255706      0
#> 685   0.3648301206      0
#> 686  -1.1231750969      0
#> 687  -0.0726600863      0
#> 688   1.0193815591      0
#> 689   0.3564575410      0
#> 690  -0.9415742959      0
#> 691   0.1839146367      0
#> 692   1.9879407476      0
#> 693   0.9984732426      0
#> 694   0.7231675008      0
#> 695   0.2738147146      0
#> 696  -2.7269827457      0
#> 697  -0.0007171029      0
#> 698  -0.8113486345      0
#> 699  -0.9683243322      0
#> 700   0.2914818560      0
#> 701  -1.9845226903      0
#> 702  -0.7494307837      0
#> 703   0.1273061045      0
#> 704   0.7863849123      0
#> 705  -0.0800875614      0
#> 706   0.8382276517      0
#> 707  -0.1890406049      0
#> 708   0.9696687867      0
#> 709   0.2291456648      0
#> 710  -0.4321940421      0
#> 711   0.0448699472      0
#> 712  -0.1884876920      0
#> 713  -1.6063303342      0
#> 714   1.1508538196      0
#> 715   1.7073534319      0
#> 716  -1.7914411948      0
#> 717  -0.5491707387      0
#> 718  -0.8200601985      0
#> 719  -1.6143855062      0
#> 720   0.2567684408      0
#> 721  -2.5346511544      0
#> 722   1.8937606764      0
#> 723  -1.8910000310      0
#> 724   0.6829127291      0
#> 725  -0.2341220180      0
#> 726  -1.5606174425      0
#> 727   0.2593604798      0
#> 728  -0.9654453579      0
#> 729  -0.4308020293      0
#> 730  -1.0993341072      0
#> 731  -1.9294112816      0
#> 732   0.1264283214      0
#> 733   0.1525207067      0
#> 734   1.0568432450      0
#> 735   1.3047440846      0
#> 736  -1.0006677541      0
#> 737  -0.4624916495      0
#> 738  -0.2782737576      0
#> 739   0.5894169549      0
#> 740   0.6737821363      0
#> 741   0.0492658003      0
#> 742   0.9316808830      0
#> 743   1.9492137132      0
#> 744   1.2164128621      0
#> 745   0.1970510799      0
#> 746   0.6400354033      0
#> 747  -1.5014935604      0
#> 748  -0.4305155049      0
#> 749   2.5712207581      0
#> 750  -0.8628419422      0
#> 751  -0.2286980546      0
#> 752   0.5722015935      0
#> 753  -0.2211060922      0
#> 754  -0.6875050487      0
#> 755   0.9686178710      0
#> 756  -0.6838961173      0
#> 757  -2.3233191125      0
#> 758   0.0086937267      0
#> 759   2.4150288367      0
#> 760   2.1219878206      0
#> 761   1.5164362564      0
#> 762  -0.0132509967      0
#> 763  -0.0878686619      0
#> 764   2.3303017229      0
#> 765  -0.7200714233      0
#> 766   1.0463875577      0
#> 767   0.6908614016      0
#> 768   0.5417680160      0
#> 769   0.4197792770      0
#> 770  -1.1137177341      0
#> 771  -2.1256446226      0
#> 772   0.4796759919      0
#> 773   0.5829578493      0
#> 774  -0.1585099416      0
#> 775   0.9810760948      0
#> 776  -0.5335834220      0
#> 777  -1.7995960961      0
#> 778   1.0001064462      0
#> 779   0.8216886890      0
#> 780  -0.2925632506      0
#> 781  -0.0436971161      0
#> 782   1.1623051942      0
#> 783  -2.8218490912      0
#> 784  -0.2509907205      0
#> 785  -1.0858459664      0
#> 786   1.8450280388      0
#> 787   1.4647134238      0
#> 788   1.0935810686      0
#> 789  -0.2960085874      0
#> 790   1.5962153574      0
#> 791   0.7728545501      0
#> 792  -0.7476775332      0
#> 793   1.2410643833      0
#> 794   1.9401925804      0
#> 795  -0.2145218754      0
#> 796   0.2316626671      0
#> 797   0.1501938949      0
#> 798   0.2850064660      0
#> 799   2.3307691989      0
#> 800  -0.2885494120      0
#> 801   0.3628398819      0
#> 802   0.4609292917      0
#> 803  -0.3874992979      0
#> 804  -0.9037870414      0
#> 805   0.2151813267      0
#> 806  -0.6880549884      0
#> 807   0.5140112289      0
#> 808   2.0942775476      0
#> 809  -0.0188926597      0
#> 810  -0.3962744531      0
#> 811  -0.2844794675      0
#> 812   0.6030666057      0
#> 813  -2.7623505053      0
#> 814  -0.2102575395      0
#> 815  -0.6062546818      0
#> 816  -1.8025817318      0
#> 817   1.1432170653      0
#> 818   0.1583367309      0
#> 819   0.2455582029      0
#> 820  -1.7885427997      0
#> 821  -1.6947179137      0
#> 822  -1.6016685124      0
#> 823  -0.4383196213      0
#> 824  -0.5227257321      0
#> 825  -0.6780361620      0
#> 826   0.6200044782      0
#> 827   0.2111326404      0
#> 828   0.2001592954      0
#> 829   1.1690215659      0
#> 830  -1.6751407540      0
#> 831   0.4772297911      0
#> 832  -0.6445359488      0
#> 833   1.0838800456      0
#> 834  -0.3977519716      0
#> 835   0.4214708853      0
#> 836  -0.4395472536      0
#> 837   0.7739690907      0
#> 838  -2.9132686264      0
#> 839   0.9436408610      0
#> 840  -0.2013905148      0
#> 841   0.6731816544      0
#> 842  -0.0015464224      0
#> 843   0.7153128876      0
#> 844  -0.5009650491      0
#> 845   0.2516185278      0
#> 846  -1.4764319976      0
#> 847  -0.0805839553      0
#> 848   1.8434151821      0
#> 849   0.8404145131      0
#> 850   0.4546430406      0
#> 851   1.4524210312      0
#> 852  -0.4798282719      0
#> 853  -0.5796774515      0
#> 854  -1.1178109134      0
#> 855   0.2480605263      0
#> 856   0.9014857580      0
#> 857  -0.1566519304      0
#> 858   1.0747778174      0
#> 859  -1.2841161739      0
#> 860  -0.4029382539      0
#> 861   0.1689093924      0
#> 862  -0.1072931416      0
#> 863   1.3728863135      0
#> 864  -0.5868736105      0
#> 865   0.2241430577      0
#> 866  -0.2913462953      0
#> 867   0.4318423477      0
#> 868   0.3348958236      0
#> 869   0.1563132827      0
#> 870  -0.1125102178      0
#> 871  -1.8363995661      0
#> 872  -1.1514315551      0
#> 873  -1.0310876392      0
#> 874  -0.6177408361      0
#> 875  -1.5642208370      0
#> 876  -0.3966178294      0
#> 877   0.1433951859      0
#> 878   0.3720447000      0
#> 879   1.7420025361      0
#> 880  -1.3357262079      0
#> 881   0.5642189699      0
#> 882  -1.0590967960      0
#> 883  -0.1641950122      0
#> 884  -0.1011982204      0
#> 885  -2.2138542137      0
#> 886   0.2392296615      0
#> 887  -1.7587643635      0
#> 888  -0.0683195913      0
#> 889   0.2833321507      0
#> 890   1.1772463963      0
#> 891   0.1431583325      0
#> 892   1.2230688964      0
#> 893   0.8152973412      0
#> 894  -0.4831468355      0
#> 895  -1.2222874934      0
#> 896  -0.5949036643      0
#> 897   0.4204386797      0
#> 898  -0.3050397634      0
#> 899  -0.1282992847      0
#> 900   0.1029157801      0
#> 901  -0.5532461196      0
#> 902  -0.2127294987      0
#> 903   0.2130008070      0
#> 904   1.2773706361      0
#> 905  -1.3942167916      0
#> 906  -0.5150044017      1
#> 907   0.7251297322      1
#> 908   0.1735460576      1
#> 909   0.4519824370      1
#> 910   0.9881722157      1
#> 911  -0.1033026638      1
#> 912   3.2254221423      1
#> 913  -0.9789725858      1
#> 914   3.0522779767      1
#> 915   1.2271063482      1
#> 916  -1.2059987720      1
#> 917   0.8497637146      1
#> 918   0.8480705621      1
#> 919   1.9756070479      1
#> 920   2.5698593141      1
#> 921   0.1737606241      1
#> 922  -0.0423249516      1
#> 923   0.5630741833      1
#> 924   1.7084340502      1
#> 925  -0.4067357884      1
#> 926   0.0678229638      1
#> 927  -0.3953953540      1
#> 928   1.2350031427      1
#> 929   0.2372776018      1
#> 930   1.3875220699      1
#> 931   1.0050327146      1
#> 932   0.6054730105      1
#> 933   1.3030206023      1
#> 934   1.8465908215      1
#> 935   0.1216080398      1
#> 936   0.3867578142      1
#> 937   1.1488481664      1
#> 938   0.5553095709      1
#> 939   1.7020373225      1
#> 940   1.8265787073      1
#> 941   0.2986798281      1
#> 942   0.8863892922      1
#> 943   2.1653400391      1
#> 944   0.1215934620      1
#> 945   0.9205136033      1
#> 946  -0.2607625250      1
#> 947  -2.0163701539      1
#> 948   0.4760055612      1
#> 949  -0.0271884721      1
#> 950   0.8758544033      1
#> 951   0.4036162805      1
#> 952  -0.1253604349      1
#> 953  -0.0115382507      1
#> 954   0.7534473600      1
#> 955   0.2399671900      1
#> 956  -0.8129533793      1
#> 957   0.2535200018      1
#> 958   0.0089759265      1
#> 959   0.1025249086      1
#> 960   1.4542601211      1
#> 961   0.7681415032      1
#> 962   2.1760115732      1
#> 963  -0.0696944135      1
#> 964   1.9754189309      1
#> 965   0.1177743477      1
#> 966   0.3295521175      1
#> 967   0.9479471051      1
#> 968   1.6624714797      1
#> 969   0.1027118753      1
#> 970   1.6467809623      1
#> 971   0.4103208589      1
#> 972   0.7648345370      1
#> 973   1.2439877042      1
#> 974   2.8802478487      1
#> 975   1.7407488170      1
#> 976   1.1373169947      1
#> 977   0.2050784824      1
#> 978   0.4201496262      1
#> 979   0.6923519008      1
#> 980   0.5101299515      1
#> 981   0.7975563411      1
#> 982   0.9291658048      1
#> 983   1.5320130373      1
#> 984   1.7309400565      1
#> 985   0.3217241668      1
#> 986   0.7654116096      1
#> 987   1.4609189130      1
#> 988   0.8901538772      1
#> 989   0.8214497917      1
#> 990   0.8078672538      1
#> 991   0.2599189908      1
#> 992   0.6929552602      1
#> 993   2.2597890064      1
#> 994   1.5477202977      1
#> 995   1.7987752012      1
#> 996  -0.1716750061      1
#> 997   1.9657058869      1
#> 998  -0.7027064150      1
#> 999   1.4062261763      1
#> 1000  1.3242880454      1