Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve.

n_samples

Number of samples to draw.

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%).

min_events

Minimum number of events required in the sample.

Value

Returns a data.frame.

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1    -1.079826e+00      0
#> 2     1.865564e+00      0
#> 3     3.078982e-01      0
#> 4     4.401127e-01      0
#> 5    -2.574729e-01      0
#> 6     5.647357e-02      0
#> 7    -9.606067e-01      0
#> 8    -2.898816e-02      0
#> 9     3.915223e-01      0
#> 10   -1.082319e+00      0
#> 11    3.658698e-01      0
#> 12   -7.685969e-01      0
#> 13    1.320428e+00      0
#> 14    5.185194e-02      0
#> 15    1.496220e-01      0
#> 16   -7.045586e-01      0
#> 17   -7.472412e-01      0
#> 18   -1.910572e+00      0
#> 19    1.683096e+00      0
#> 20    5.352485e-01      0
#> 21    1.318660e+00      0
#> 22    8.802876e-01      0
#> 23    3.777308e-01      0
#> 24   -1.654817e+00      0
#> 25   -1.874423e+00      0
#> 26   -1.083886e+00      0
#> 27    1.122137e-02      0
#> 28    1.374290e-01      0
#> 29   -1.533788e-01      0
#> 30   -3.530628e-01      0
#> 31   -1.229439e+00      0
#> 32   -5.820800e-01      0
#> 33    5.369865e-01      0
#> 34   -8.731431e-01      0
#> 35   -1.245701e+00      0
#> 36   -7.986329e-01      0
#> 37   -1.418992e+00      0
#> 38   -2.466413e-01      0
#> 39   -4.561048e-01      0
#> 40    1.925027e+00      0
#> 41   -9.286013e-01      0
#> 42   -7.558660e-01      0
#> 43    1.228138e+00      0
#> 44    1.492506e+00      0
#> 45    9.170466e-01      0
#> 46   -2.237015e-01      0
#> 47   -4.416210e-01      0
#> 48   -8.658978e-01      0
#> 49    7.363383e-01      0
#> 50    7.770239e-01      0
#> 51    3.417820e-01      0
#> 52    5.341114e-01      0
#> 53   -1.625926e+00      0
#> 54    1.645930e+00      0
#> 55    1.005367e+00      0
#> 56   -2.834536e-01      0
#> 57    1.285665e-01      0
#> 58    5.413718e-01      0
#> 59   -4.751499e-01      0
#> 60    7.590304e-02      0
#> 61   -1.254337e+00      0
#> 62    1.031871e-01      0
#> 63   -3.968495e-01      0
#> 64   -3.646399e-02      0
#> 65   -1.602450e+00      0
#> 66   -6.338296e-01      0
#> 67   -7.805577e-01      0
#> 68   -3.208731e-02      0
#> 69    1.527444e-01      0
#> 70    1.157519e+00      0
#> 71   -3.549879e-01      0
#> 72   -4.628662e-01      0
#> 73   -1.073898e+00      0
#> 74   -3.551089e-01      0
#> 75    1.525003e+00      0
#> 76    1.046266e+00      0
#> 77    2.311688e-01      0
#> 78    2.550695e-01      0
#> 79    4.867672e-01      0
#> 80   -2.105689e-01      0
#> 81    4.818680e-01      0
#> 82    1.357019e+00      0
#> 83   -1.159401e+00      0
#> 84    1.084409e+00      0
#> 85    3.048460e-01      0
#> 86    1.961942e+00      0
#> 87   -1.116454e-01      0
#> 88   -3.187917e-01      0
#> 89    3.115980e-01      0
#> 90    1.030966e+00      0
#> 91   -1.713788e+00      0
#> 92    1.019517e-01      0
#> 93    9.803652e-01      0
#> 94    1.933972e+00      0
#> 95    1.612916e+00      0
#> 96   -1.982832e+00      0
#> 97   -1.666133e+00      0
#> 98    1.002048e+00      0
#> 99    5.839814e-02      0
#> 100  -1.242239e+00      0
#> 101  -4.591109e-01      0
#> 102   6.246299e-02      0
#> 103  -1.350452e+00      0
#> 104   4.480144e-01      0
#> 105  -3.608687e-01      0
#> 106   2.598053e-01      0
#> 107  -1.299970e-01      0
#> 108  -4.318741e-01      0
#> 109   1.443369e+00      0
#> 110   1.099337e+00      0
#> 111   8.953756e-01      0
#> 112  -2.142394e-01      0
#> 113   4.586038e-01      0
#> 114  -3.572908e-01      0
#> 115   8.718580e-01      0
#> 116  -2.360350e-01      0
#> 117  -1.790691e-01      0
#> 118  -2.262584e+00      0
#> 119   9.881467e-01      0
#> 120  -6.549068e-01      0
#> 121   9.444782e-01      0
#> 122  -1.970057e+00      0
#> 123  -1.125815e+00      0
#> 124   4.802431e-01      0
#> 125  -1.270992e-01      0
#> 126   1.629365e+00      0
#> 127   6.719415e-01      0
#> 128  -1.739850e+00      0
#> 129  -1.361387e-01      0
#> 130   9.235620e-01      0
#> 131   8.173088e-02      0
#> 132  -8.382205e-01      0
#> 133   9.067711e-03      0
#> 134  -4.126390e-01      0
#> 135  -9.373899e-01      0
#> 136   8.225236e-01      0
#> 137  -2.376440e+00      0
#> 138  -7.049333e-01      0
#> 139   7.497004e-01      0
#> 140  -7.934171e-01      0
#> 141   3.699536e-01      0
#> 142   7.179194e-01      0
#> 143   7.592572e-02      0
#> 144  -4.541801e-02      0
#> 145  -1.819613e+00      0
#> 146   4.497768e-01      0
#> 147   8.211045e-01      0
#> 148  -2.333236e-01      0
#> 149  -6.758334e-01      0
#> 150  -8.743819e-01      0
#> 151   2.441137e-01      0
#> 152  -5.822278e-01      0
#> 153   8.705117e-01      0
#> 154   1.617422e+00      0
#> 155   7.758860e-01      0
#> 156  -4.007756e-01      0
#> 157  -1.800067e+00      0
#> 158  -3.988761e-02      0
#> 159  -3.029957e+00      0
#> 160   1.019497e+00      0
#> 161   1.104013e+00      0
#> 162   7.585783e-01      0
#> 163  -7.473717e-01      0
#> 164   4.265469e-01      0
#> 165  -1.484597e+00      0
#> 166  -5.474911e-01      0
#> 167   3.857214e-01      0
#> 168   9.044388e-01      0
#> 169  -1.486476e+00      0
#> 170   3.002692e-01      0
#> 171  -6.141717e-01      0
#> 172  -1.417303e+00      0
#> 173   4.300499e-01      0
#> 174  -1.613950e+00      0
#> 175  -1.797745e-01      0
#> 176   6.216724e-01      0
#> 177   3.142082e-02      0
#> 178   6.125913e-01      0
#> 179  -1.193889e+00      0
#> 180  -6.759013e-01      0
#> 181  -6.992631e-02      0
#> 182  -6.294415e-01      0
#> 183   1.652710e+00      0
#> 184   2.593294e-01      0
#> 185   3.425916e-01      0
#> 186   8.699276e-01      0
#> 187   5.151472e-01      0
#> 188   6.909395e-01      0
#> 189  -4.286082e-01      0
#> 190  -1.436499e+00      0
#> 191  -1.275468e+00      0
#> 192   7.111451e-01      0
#> 193   2.561302e-01      0
#> 194   1.116163e-01      0
#> 195  -3.478097e-02      0
#> 196   6.013428e-01      0
#> 197   1.320369e+00      0
#> 198  -2.079810e+00      0
#> 199   1.786900e+00      0
#> 200   1.104104e+00      0
#> 201  -2.624156e-01      0
#> 202   6.851875e-01      0
#> 203  -1.913767e+00      0
#> 204  -6.188881e-01      0
#> 205  -6.027259e-01      0
#> 206  -4.556612e-01      0
#> 207   4.505907e-02      0
#> 208  -6.110532e-01      0
#> 209   6.101515e-01      0
#> 210   3.706859e-01      0
#> 211  -5.681085e-01      0
#> 212  -7.054709e-01      0
#> 213  -4.779993e-01      0
#> 214   1.037392e-01      0
#> 215   9.721467e-01      0
#> 216  -1.066711e+00      0
#> 217   6.879693e-01      0
#> 218   6.695100e-01      0
#> 219   3.532398e-05      0
#> 220   6.677754e-01      0
#> 221  -1.248731e+00      0
#> 222   2.580201e+00      0
#> 223   4.725816e-01      0
#> 224   1.007906e+00      0
#> 225  -9.855274e-01      0
#> 226  -4.390991e-02      0
#> 227   8.465924e-02      0
#> 228  -6.045467e-01      0
#> 229   2.662125e+00      0
#> 230   8.385686e-01      0
#> 231   3.839982e-01      0
#> 232  -1.056051e+00      0
#> 233  -4.134286e-01      0
#> 234   7.822376e-01      0
#> 235   2.822166e-01      0
#> 236  -4.604888e-01      0
#> 237   4.148585e-01      0
#> 238  -1.448375e-01      0
#> 239   8.678424e-01      0
#> 240  -1.289350e+00      0
#> 241   7.168998e-01      0
#> 242  -1.297085e+00      0
#> 243   5.544360e-01      0
#> 244  -1.116243e-02      0
#> 245  -1.119087e+00      0
#> 246   9.319134e-01      0
#> 247   1.965980e+00      0
#> 248   2.035467e-01      0
#> 249   1.327297e+00      0
#> 250  -5.442197e-02      0
#> 251  -4.685391e-01      0
#> 252  -4.590322e-01      0
#> 253   3.810252e-01      0
#> 254   9.344889e-01      0
#> 255   1.206370e+00      0
#> 256   5.610666e-01      0
#> 257   5.510561e-01      0
#> 258  -5.901358e-01      0
#> 259  -1.495692e-01      0
#> 260  -9.616293e-01      0
#> 261   4.852717e-01      0
#> 262   7.143644e-01      0
#> 263   1.229434e-01      0
#> 264  -3.615064e-01      0
#> 265   4.761906e-01      0
#> 266   1.134976e+00      0
#> 267   1.035121e+00      0
#> 268   1.669154e-01      0
#> 269   1.501471e-01      0
#> 270   1.734980e+00      0
#> 271   6.660168e-01      0
#> 272   1.946961e-01      0
#> 273   8.585831e-01      0
#> 274  -1.928282e-01      0
#> 275   1.109761e+00      0
#> 276  -1.947358e-02      0
#> 277   1.038813e+00      0
#> 278   2.215080e-01      0
#> 279  -1.697753e+00      0
#> 280  -1.464104e+00      0
#> 281   7.218501e-01      0
#> 282  -1.167467e+00      0
#> 283  -9.020323e-01      0
#> 284  -1.305014e+00      0
#> 285   1.053392e-01      0
#> 286   9.546514e-01      0
#> 287   3.835307e-01      0
#> 288   5.674394e-01      0
#> 289   3.765447e-01      0
#> 290   4.309005e-01      0
#> 291   1.246626e+00      0
#> 292   2.649445e-03      0
#> 293   6.559641e-01      0
#> 294  -8.809996e-01      0
#> 295   1.623387e+00      0
#> 296  -8.741230e-01      0
#> 297   5.876426e-01      0
#> 298   8.631180e-01      0
#> 299  -2.160495e-01      0
#> 300  -1.543829e+00      0
#> 301  -1.991246e+00      0
#> 302  -3.819118e-01      0
#> 303   8.763099e-01      0
#> 304  -1.829659e-01      0
#> 305  -8.045158e-01      0
#> 306   2.775432e-01      0
#> 307  -1.268220e+00      0
#> 308   7.794117e-02      0
#> 309   4.141816e-01      0
#> 310  -1.257135e+00      0
#> 311   1.183636e-02      0
#> 312   1.039098e+00      0
#> 313  -1.680311e+00      0
#> 314  -6.181671e-01      0
#> 315   9.586018e-01      0
#> 316   6.126609e-01      0
#> 317  -6.642312e-01      0
#> 318  -1.202867e+00      0
#> 319  -1.055433e+00      0
#> 320   1.451699e-02      0
#> 321  -1.307633e+00      0
#> 322   2.129321e-01      0
#> 323  -1.852051e+00      0
#> 324   1.203765e+00      0
#> 325   2.755526e-01      0
#> 326   3.990979e-01      0
#> 327  -1.130178e+00      0
#> 328  -1.544203e+00      0
#> 329   6.534346e-01      0
#> 330  -1.310965e+00      0
#> 331  -1.276825e-01      0
#> 332   1.630143e+00      0
#> 333  -3.876742e-01      0
#> 334   3.661484e-01      0
#> 335  -3.441405e-01      0
#> 336  -1.515589e+00      0
#> 337   9.789628e-01      0
#> 338   8.313988e-03      0
#> 339   1.890435e+00      0
#> 340  -2.781179e-01      0
#> 341  -1.182166e+00      0
#> 342  -1.953347e-01      0
#> 343  -7.663596e-01      0
#> 344  -5.373876e-01      0
#> 345   1.191002e+00      0
#> 346  -5.194650e-01      0
#> 347  -1.126156e+00      0
#> 348  -1.587790e+00      0
#> 349   1.490836e+00      0
#> 350  -7.252538e-01      0
#> 351  -2.683543e-01      0
#> 352  -1.021608e+00      0
#> 353  -1.220490e+00      0
#> 354  -3.785751e-01      0
#> 355  -2.485099e-01      0
#> 356   3.563759e-01      0
#> 357   4.378833e-01      0
#> 358  -1.642950e+00      0
#> 359   4.591906e-01      0
#> 360  -1.702184e-01      0
#> 361  -2.615325e-01      0
#> 362  -1.466392e-02      0
#> 363  -1.640982e+00      0
#> 364   4.664166e-02      0
#> 365  -1.044971e-01      0
#> 366   1.087799e+00      0
#> 367   4.559518e-01      0
#> 368  -9.969053e-01      0
#> 369   1.463876e-01      0
#> 370  -1.458199e+00      0
#> 371   9.187088e-01      0
#> 372  -4.510749e-01      0
#> 373  -1.002164e+00      0
#> 374   1.675754e+00      0
#> 375  -1.328918e+00      0
#> 376  -1.795249e-01      0
#> 377   1.016137e+00      0
#> 378  -1.768296e-01      0
#> 379  -4.657859e-01      0
#> 380  -2.341052e+00      0
#> 381   6.817318e-01      0
#> 382  -3.088169e-01      0
#> 383  -1.270650e+00      0
#> 384  -3.223903e-01      0
#> 385  -2.794107e-01      0
#> 386  -6.527186e-01      0
#> 387  -2.741663e-01      0
#> 388   1.349090e+00      0
#> 389   1.993125e-01      0
#> 390  -1.509535e-01      0
#> 391  -5.737442e-01      0
#> 392  -2.276951e+00      0
#> 393   1.464094e+00      0
#> 394   9.110835e-03      0
#> 395   1.512895e+00      0
#> 396  -2.973183e-01      0
#> 397   1.706860e-01      0
#> 398  -1.249852e-01      0
#> 399   8.753567e-01      0
#> 400  -1.098057e+00      0
#> 401  -7.931864e-01      0
#> 402  -1.183945e+00      0
#> 403   3.347109e-01      0
#> 404   1.032647e+00      0
#> 405   1.735118e-02      0
#> 406  -1.134371e+00      0
#> 407   8.828372e-01      0
#> 408   2.436672e-01      0
#> 409  -5.535534e-01      0
#> 410   2.614688e-01      0
#> 411   1.799972e+00      0
#> 412  -1.489532e+00      0
#> 413   2.878525e-01      0
#> 414   1.554029e-01      0
#> 415   7.993543e-01      0
#> 416  -3.794677e-01      0
#> 417   4.029015e-01      0
#> 418  -1.222378e+00      0
#> 419  -9.212114e-01      0
#> 420  -5.470433e-01      0
#> 421  -7.328691e-01      0
#> 422  -1.235061e+00      0
#> 423  -4.473106e-01      0
#> 424  -7.867507e-02      0
#> 425  -6.055189e-01      0
#> 426   7.271534e-01      0
#> 427   1.066027e+00      0
#> 428   3.372821e-01      0
#> 429   1.104453e+00      0
#> 430  -2.012836e-01      0
#> 431   1.846774e+00      0
#> 432   2.891225e-01      0
#> 433   2.528460e-01      0
#> 434  -8.925726e-01      0
#> 435   4.772199e-01      0
#> 436   1.368086e+00      0
#> 437  -1.261296e+00      0
#> 438  -1.575112e-01      0
#> 439   1.672357e+00      0
#> 440  -2.247552e-01      0
#> 441   6.284369e-01      0
#> 442   1.367516e+00      0
#> 443  -2.673889e-01      0
#> 444  -5.718630e-01      0
#> 445   1.573933e-01      0
#> 446  -3.338049e-01      0
#> 447  -5.192858e-01      0
#> 448   1.757077e+00      0
#> 449  -1.196371e+00      0
#> 450  -6.497206e-01      0
#> 451   8.694628e-02      0
#> 452   4.204782e-02      0
#> 453  -9.903370e-01      0
#> 454  -1.500603e+00      0
#> 455  -7.998123e-01      0
#> 456   3.350372e-01      0
#> 457  -1.924002e-01      0
#> 458  -1.296005e+00      0
#> 459   8.055396e-01      0
#> 460   4.447461e-01      0
#> 461   1.063883e+00      0
#> 462  -5.553995e-01      0
#> 463  -2.701596e-01      0
#> 464   4.425560e-02      0
#> 465   1.174473e+00      0
#> 466  -4.380947e-01      0
#> 467  -1.559739e+00      0
#> 468  -1.137259e+00      0
#> 469  -3.727662e-01      0
#> 470   1.613908e-01      0
#> 471  -8.647406e-01      0
#> 472  -1.479097e+00      0
#> 473   1.399286e+00      0
#> 474  -8.607908e-01      0
#> 475   1.510003e-02      0
#> 476   6.159297e-01      0
#> 477   1.527986e-01      0
#> 478   1.518655e-01      0
#> 479  -1.926066e-02      0
#> 480  -1.227310e+00      0
#> 481  -1.012152e+00      0
#> 482  -5.461762e-02      0
#> 483  -1.752248e-02      0
#> 484   1.445802e+00      0
#> 485  -9.987579e-01      0
#> 486   3.347130e-01      0
#> 487  -9.419508e-01      0
#> 488  -1.029145e-01      0
#> 489  -9.332238e-02      0
#> 490  -8.388680e-01      0
#> 491   1.161761e+00      0
#> 492  -4.688665e-01      0
#> 493   2.741559e-01      0
#> 494  -6.714349e-01      0
#> 495   5.471271e-01      0
#> 496  -1.636486e+00      0
#> 497   1.465677e-01      0
#> 498  -1.725550e-01      0
#> 499   5.258527e-01      0
#> 500  -6.647500e-01      0
#> 501  -1.136982e+00      0
#> 502  -9.359911e-01      0
#> 503  -1.801822e-01      0
#> 504  -7.995143e-01      0
#> 505  -2.476808e-01      0
#> 506  -8.588268e-01      0
#> 507   1.249159e+00      0
#> 508   2.838277e-01      0
#> 509  -1.106623e-01      0
#> 510  -3.481193e-01      0
#> 511   5.942690e-01      0
#> 512  -8.080769e-01      0
#> 513   3.961721e-02      0
#> 514  -5.884255e-01      0
#> 515   1.411643e-02      0
#> 516   3.754300e-01      0
#> 517  -7.496982e-02      0
#> 518   8.232815e-01      0
#> 519   6.131294e-02      0
#> 520   2.111602e+00      0
#> 521  -3.884675e-01      0
#> 522   1.346570e+00      0
#> 523  -2.938103e-01      0
#> 524  -1.829069e+00      0
#> 525   2.869385e-01      0
#> 526   2.130226e-01      0
#> 527  -1.407560e+00      0
#> 528   8.215047e-02      0
#> 529   2.300114e+00      0
#> 530  -1.287704e+00      0
#> 531  -3.847893e-01      0
#> 532   1.172066e+00      0
#> 533  -5.698792e-02      0
#> 534  -4.707387e-01      0
#> 535   7.582339e-01      0
#> 536   1.022376e+00      0
#> 537  -1.670548e+00      0
#> 538  -1.185412e+00      0
#> 539  -7.325173e-01      0
#> 540  -7.836180e-01      0
#> 541   9.441970e-01      0
#> 542  -4.690581e-01      0
#> 543   1.186833e+00      0
#> 544  -1.284444e+00      0
#> 545  -1.228471e+00      0
#> 546   5.831471e-01      0
#> 547   4.388796e-01      0
#> 548  -3.012700e-02      0
#> 549   2.861290e-01      0
#> 550   1.234456e+00      0
#> 551   6.044322e-01      0
#> 552  -2.802808e-01      0
#> 553  -7.838740e-01      0
#> 554  -6.356035e-01      0
#> 555  -9.749369e-01      0
#> 556  -1.628249e+00      0
#> 557  -4.189779e-01      0
#> 558  -1.172275e-01      0
#> 559  -1.048606e-02      0
#> 560  -1.786613e+00      0
#> 561   8.511369e-01      0
#> 562   9.125720e-02      0
#> 563  -6.962000e-01      0
#> 564   1.063195e+00      0
#> 565   8.828338e-01      0
#> 566   3.469859e-01      0
#> 567   5.319887e-01      0
#> 568  -3.735497e-01      0
#> 569  -1.184531e+00      0
#> 570   1.560298e+00      0
#> 571   2.989032e-01      0
#> 572   1.642929e-02      0
#> 573  -8.930677e-01      0
#> 574   2.384702e-01      0
#> 575   1.357289e-01      0
#> 576  -1.627451e+00      0
#> 577   8.133428e-01      0
#> 578   1.591710e+00      0
#> 579   7.342008e-01      0
#> 580  -1.943932e+00      0
#> 581   2.106866e-01      0
#> 582  -6.240644e-01      0
#> 583   1.624343e+00      0
#> 584   2.600987e+00      0
#> 585  -1.400786e-01      0
#> 586   4.393670e-01      0
#> 587   1.496896e-01      0
#> 588   1.000730e+00      0
#> 589   4.359134e-01      0
#> 590   8.730846e-01      0
#> 591  -2.095689e+00      0
#> 592   8.563411e-01      0
#> 593  -1.230801e+00      0
#> 594   6.069694e-01      0
#> 595   2.328512e-01      0
#> 596   1.545855e+00      0
#> 597   6.196110e-03      0
#> 598  -7.146895e-02      0
#> 599  -8.519948e-01      0
#> 600  -2.874672e-01      0
#> 601  -7.439484e-01      0
#> 602   8.475742e-01      0
#> 603   6.486524e-01      0
#> 604   1.097868e-01      0
#> 605  -2.640108e-01      0
#> 606  -1.060835e+00      0
#> 607  -8.721873e-01      0
#> 608  -2.530046e+00      0
#> 609  -2.339699e+00      0
#> 610  -1.239977e-01      0
#> 611   7.349534e-01      0
#> 612   3.214342e-01      0
#> 613   8.241475e-01      0
#> 614  -1.486570e-01      0
#> 615   9.916793e-01      0
#> 616   1.729517e+00      0
#> 617   9.197668e-02      0
#> 618   7.752594e-01      0
#> 619  -5.741457e-02      0
#> 620   4.812932e-01      0
#> 621   1.468052e+00      0
#> 622   9.557398e-01      0
#> 623  -2.028828e+00      0
#> 624   1.936969e+00      0
#> 625   1.095745e+00      0
#> 626  -2.194288e-01      0
#> 627  -4.247732e-01      0
#> 628   9.768198e-01      0
#> 629  -1.373098e+00      0
#> 630  -4.128966e-01      0
#> 631  -7.174079e-01      0
#> 632   1.157147e+00      0
#> 633   2.412284e+00      0
#> 634   3.038530e-01      0
#> 635   2.001635e+00      0
#> 636   1.222059e+00      0
#> 637  -8.997375e-01      0
#> 638   6.576092e-01      0
#> 639  -1.831438e+00      0
#> 640   1.171125e+00      0
#> 641   3.020613e-02      0
#> 642   8.465707e-02      0
#> 643   1.769278e+00      0
#> 644   7.607681e-01      0
#> 645  -5.172397e-01      0
#> 646   1.348445e+00      0
#> 647  -2.240542e-01      0
#> 648   1.960056e+00      0
#> 649   1.787338e-01      0
#> 650  -5.672362e-02      0
#> 651  -1.104735e+00      0
#> 652  -4.417685e-01      0
#> 653  -1.261480e+00      0
#> 654  -5.890367e-01      0
#> 655   5.340296e-01      0
#> 656  -6.531175e-01      0
#> 657   2.355078e-01      0
#> 658  -5.787619e-01      0
#> 659   6.028702e-01      0
#> 660   2.156330e-01      0
#> 661   3.121042e-01      0
#> 662   1.261384e-01      0
#> 663  -6.759369e-01      0
#> 664  -6.245730e-01      0
#> 665  -1.557180e+00      0
#> 666  -1.531338e+00      0
#> 667   1.098340e+00      0
#> 668  -1.293605e+00      0
#> 669  -1.207961e+00      0
#> 670   2.458513e-01      0
#> 671  -1.081688e+00      0
#> 672  -1.571890e-01      0
#> 673   2.157280e-01      0
#> 674   2.021543e-01      0
#> 675  -1.663835e+00      0
#> 676  -1.058820e+00      0
#> 677  -1.650640e+00      0
#> 678  -7.362403e-01      0
#> 679   2.107216e-01      0
#> 680   8.965106e-01      0
#> 681  -7.717545e-01      0
#> 682  -1.860030e+00      0
#> 683  -6.814578e-01      0
#> 684  -3.317515e-01      0
#> 685   1.088089e+00      0
#> 686  -5.411987e-01      0
#> 687  -5.585099e-01      0
#> 688   5.425597e-01      0
#> 689   4.826610e-01      0
#> 690  -1.194152e+00      0
#> 691   2.624522e+00      0
#> 692  -8.794991e-01      0
#> 693  -5.730628e-02      0
#> 694   4.959515e-01      0
#> 695   2.677956e-01      0
#> 696   8.234512e-01      0
#> 697  -1.197064e+00      0
#> 698  -1.286194e+00      0
#> 699   3.974644e-01      0
#> 700   3.529625e-01      0
#> 701  -6.202418e-01      0
#> 702  -1.640553e+00      0
#> 703   1.411247e-01      0
#> 704   1.615384e+00      0
#> 705   8.639294e-01      0
#> 706  -3.716800e-01      0
#> 707   3.256891e-01      0
#> 708   9.274675e-01      0
#> 709  -1.249283e+00      0
#> 710  -1.583320e+00      0
#> 711   1.766247e-01      0
#> 712   9.819473e-01      0
#> 713  -2.090143e+00      0
#> 714   7.172983e-01      0
#> 715  -4.986074e-01      0
#> 716  -1.026888e+00      0
#> 717   4.439318e-01      0
#> 718   1.427217e+00      0
#> 719  -5.487240e-02      0
#> 720   4.262868e-01      0
#> 721  -2.666164e-01      0
#> 722  -1.224298e+00      0
#> 723   5.809780e-01      0
#> 724  -1.471491e+00      0
#> 725  -7.057402e-01      0
#> 726   1.139112e+00      0
#> 727  -1.187973e+00      0
#> 728   2.330121e-01      0
#> 729   9.452495e-01      0
#> 730  -4.289537e-01      0
#> 731  -1.235838e+00      0
#> 732  -1.147617e+00      0
#> 733   9.811300e-01      0
#> 734  -6.950346e-01      0
#> 735   3.245935e-01      0
#> 736  -1.046480e+00      0
#> 737   6.382538e-01      0
#> 738   1.111802e+00      0
#> 739   1.131837e+00      0
#> 740  -9.523528e-01      0
#> 741   2.343299e-01      0
#> 742   6.114530e-01      0
#> 743   3.148865e-01      0
#> 744   1.057100e+00      0
#> 745  -1.546251e-01      0
#> 746   2.153249e-01      0
#> 747   9.982214e-01      0
#> 748   1.259636e+00      0
#> 749  -1.341921e-01      0
#> 750   9.141539e-01      0
#> 751  -3.791369e-01      0
#> 752   1.773838e+00      0
#> 753   7.163837e-01      0
#> 754  -2.031592e+00      0
#> 755   3.485513e-02      0
#> 756   1.360915e+00      0
#> 757  -1.546290e+00      0
#> 758  -1.763512e+00      0
#> 759  -2.977883e-01      0
#> 760   3.533547e-01      0
#> 761   3.890898e-02      0
#> 762   8.450224e-01      0
#> 763  -3.055273e-01      0
#> 764   2.140512e+00      0
#> 765   2.835606e-01      0
#> 766  -5.323255e-01      0
#> 767   9.688362e-01      0
#> 768  -2.474319e+00      0
#> 769  -1.405564e+00      0
#> 770   2.036801e-01      0
#> 771   1.691483e-01      0
#> 772  -1.417097e-01      0
#> 773   4.280244e-01      0
#> 774  -1.523955e+00      0
#> 775  -1.190864e+00      0
#> 776  -2.126854e-01      0
#> 777   8.025094e-01      0
#> 778  -1.492857e+00      0
#> 779   1.821238e+00      0
#> 780   1.335851e+00      0
#> 781   7.217197e-01      0
#> 782  -1.207512e+00      0
#> 783   3.929821e-01      0
#> 784   2.833381e-02      0
#> 785  -1.407463e+00      0
#> 786   1.119176e+00      0
#> 787  -4.093132e-01      0
#> 788  -1.403194e+00      0
#> 789   2.471367e+00      0
#> 790   3.889627e-01      0
#> 791  -5.498369e-01      0
#> 792  -7.760996e-02      0
#> 793   1.922309e+00      0
#> 794   2.892511e+00      0
#> 795  -5.185406e-01      0
#> 796  -6.381449e-01      0
#> 797   8.777759e-01      0
#> 798  -2.064733e-01      0
#> 799  -7.144019e-01      0
#> 800  -1.579531e+00      0
#> 801  -1.096901e+00      0
#> 802  -7.602791e-01      0
#> 803  -4.446056e-01      0
#> 804  -2.093144e-01      0
#> 805  -1.503346e-01      0
#> 806   6.383466e-01      0
#> 807   2.429891e-01      0
#> 808  -5.254014e-01      0
#> 809   4.761558e-01      0
#> 810   2.318408e-02      0
#> 811  -1.867168e-01      0
#> 812  -5.024756e-01      0
#> 813   6.317137e-01      0
#> 814  -1.740522e-01      0
#> 815  -2.409442e-01      0
#> 816   3.835481e-01      0
#> 817   1.093887e+00      0
#> 818  -2.038521e+00      0
#> 819  -5.469220e-01      0
#> 820   7.979891e-02      0
#> 821  -9.580064e-03      0
#> 822   1.726919e-01      0
#> 823   5.600902e-01      0
#> 824   4.978459e-01      0
#> 825  -1.233621e+00      0
#> 826   1.316526e-01      0
#> 827  -2.498084e-02      0
#> 828   5.892675e-01      0
#> 829  -5.654129e-02      0
#> 830  -7.747013e-01      0
#> 831   5.407343e-01      0
#> 832   4.162375e-01      0
#> 833  -7.910290e-02      0
#> 834   5.891227e-01      0
#> 835  -2.314187e-01      0
#> 836  -4.400339e-01      0
#> 837   9.890567e-01      0
#> 838  -7.986315e-01      0
#> 839  -6.222183e-01      0
#> 840  -7.781736e-01      0
#> 841   9.682648e-01      0
#> 842  -3.521382e-01      0
#> 843   6.536943e-01      0
#> 844  -2.089576e+00      0
#> 845   1.215694e-02      0
#> 846  -6.823943e-01      0
#> 847   4.098128e-01      0
#> 848  -1.070591e+00      0
#> 849   3.447136e-01      0
#> 850   5.948587e-01      0
#> 851   3.302823e-01      0
#> 852   7.406203e-01      0
#> 853   5.654458e-01      0
#> 854   7.117215e-01      0
#> 855  -2.290261e+00      0
#> 856   1.459959e+00      0
#> 857  -1.186580e+00      0
#> 858  -2.169504e-01      0
#> 859   5.885153e-01      0
#> 860   4.613758e-01      0
#> 861  -7.735020e-01      0
#> 862  -1.011914e-01      0
#> 863   1.354436e+00      0
#> 864  -1.111635e+00      0
#> 865  -1.520406e+00      0
#> 866   2.693716e-01      0
#> 867   4.096155e-02      0
#> 868   5.902644e-01      0
#> 869  -3.147808e-01      0
#> 870   1.456952e+00      0
#> 871  -1.469339e+00      0
#> 872   1.784359e+00      0
#> 873  -1.206732e+00      0
#> 874   7.333035e-01      0
#> 875   1.888409e+00      0
#> 876  -1.526524e+00      0
#> 877  -1.162725e+00      0
#> 878  -1.073085e+00      0
#> 879   1.060550e+00      0
#> 880  -1.741245e+00      0
#> 881   1.963565e-01      0
#> 882  -1.028678e+00      0
#> 883   1.121037e+00      0
#> 884   3.436468e-02      0
#> 885   1.566269e+00      0
#> 886  -1.650010e+00      0
#> 887   1.438711e+00      1
#> 888   1.643958e+00      1
#> 889   1.814148e+00      1
#> 890  -3.263022e-01      1
#> 891   6.346464e-01      1
#> 892  -4.005543e-01      1
#> 893   1.189922e+00      1
#> 894   5.787527e-01      1
#> 895   1.701888e+00      1
#> 896   7.038143e-01      1
#> 897   4.209861e+00      1
#> 898  -7.946624e-01      1
#> 899   2.700637e-01      1
#> 900   2.305705e+00      1
#> 901   6.084988e-01      1
#> 902   3.211500e-01      1
#> 903  -3.388225e-01      1
#> 904  -4.353736e-01      1
#> 905   1.995897e+00      1
#> 906   3.962874e-01      1
#> 907   8.037286e-01      1
#> 908   4.080245e-01      1
#> 909  -3.237502e-01      1
#> 910   5.005415e-01      1
#> 911   1.110999e+00      1
#> 912   1.278565e+00      1
#> 913   1.429873e+00      1
#> 914   1.530506e+00      1
#> 915   5.523470e-01      1
#> 916   1.127292e+00      1
#> 917   1.638709e+00      1
#> 918  -8.215211e-01      1
#> 919   3.368082e-01      1
#> 920   1.544157e+00      1
#> 921   1.290424e+00      1
#> 922   3.287725e-01      1
#> 923   2.181452e+00      1
#> 924   1.347422e+00      1
#> 925   5.178683e-01      1
#> 926   9.853773e-01      1
#> 927   1.757661e-01      1
#> 928  -1.097789e+00      1
#> 929   1.482739e+00      1
#> 930   8.673324e-01      1
#> 931   1.084342e+00      1
#> 932   1.168010e+00      1
#> 933  -5.671695e-01      1
#> 934   1.565925e+00      1
#> 935   1.406999e+00      1
#> 936   8.809961e-01      1
#> 937   5.969373e-01      1
#> 938   7.801237e-01      1
#> 939   1.022682e+00      1
#> 940   5.003179e-01      1
#> 941   2.459138e+00      1
#> 942   8.191220e-01      1
#> 943   8.860514e-01      1
#> 944   2.725846e-01      1
#> 945  -2.007945e-01      1
#> 946   2.371263e-01      1
#> 947  -2.101304e-02      1
#> 948   2.582882e+00      1
#> 949   6.851567e-02      1
#> 950   4.914616e-01      1
#> 951   5.997351e-01      1
#> 952   3.065294e+00      1
#> 953  -3.353258e-01      1
#> 954  -2.706547e-01      1
#> 955   9.448887e-01      1
#> 956   5.852958e-01      1
#> 957   1.588684e+00      1
#> 958  -2.078261e-01      1
#> 959   7.717548e-01      1
#> 960   1.105897e+00      1
#> 961  -1.270205e+00      1
#> 962  -1.575177e+00      1
#> 963   5.553369e-01      1
#> 964   2.711627e+00      1
#> 965   1.632337e+00      1
#> 966  -7.843045e-01      1
#> 967  -1.432384e-01      1
#> 968   5.093697e-02      1
#> 969   1.803683e+00      1
#> 970   1.020676e+00      1
#> 971  -7.354532e-01      1
#> 972   2.886057e+00      1
#> 973   1.308724e+00      1
#> 974  -3.304214e-01      1
#> 975   9.028208e-01      1
#> 976   1.431995e+00      1
#> 977   5.481955e-01      1
#> 978   1.346232e+00      1
#> 979  -2.211339e-01      1
#> 980   2.813823e+00      1
#> 981   7.115070e-01      1
#> 982   2.819740e-01      1
#> 983   1.332566e+00      1
#> 984   2.248900e-01      1
#> 985   2.902724e-01      1
#> 986   1.771239e+00      1
#> 987   9.095103e-01      1
#> 988  -4.119520e-01      1
#> 989   6.095378e-02      1
#> 990  -1.484101e+00      1
#> 991  -3.437136e-01      1
#> 992   8.112770e-01      1
#> 993  -4.926668e-01      1
#> 994   9.097470e-01      1
#> 995   7.654180e-01      1
#> 996   6.908372e-01      1
#> 997   1.354171e-01      1
#> 998   1.336516e+00      1
#> 999   1.998585e+00      1
#> 1000  1.784064e-01      1