Samples data for a prediction model with a specified AUC and prevalence.

Samples data for a prediction model with a specified AUC and prevalence.

Usage

get_sample(auc, n_samples, prevalence, min_events = 0)

Arguments

auc

The Area Under the (receiver operating characteristic) Curve

n_samples

Number of samples to draw

prevalence

Prevalence or event rate of the binary outcome as a proportion (0.1 = 10%)

min_events

Minimum number of events required in the sample

Value

data.frame

Examples

get_sample(0.7, 1000, 0.1)
#>                  x actual
#> 1     8.650521e-01      0
#> 2    -1.319014e+00      0
#> 3     6.042113e-01      0
#> 4     1.388551e-02      0
#> 5    -1.511866e+00      0
#> 6    -5.344062e-01      0
#> 7    -1.429754e+00      0
#> 8    -3.279170e-01      0
#> 9     1.257381e+00      0
#> 10    1.120883e+00      0
#> 11    4.703352e-01      0
#> 12    1.651201e-01      0
#> 13   -1.187700e+00      0
#> 14   -3.330151e-01      0
#> 15   -5.402981e-01      0
#> 16    5.439627e-01      0
#> 17   -1.093328e-01      0
#> 18    1.511152e+00      0
#> 19   -1.984808e+00      0
#> 20    7.589802e-01      0
#> 21   -1.188699e+00      0
#> 22    2.897339e-03      0
#> 23   -8.870891e-01      0
#> 24    1.023021e+00      0
#> 25    7.934954e-01      0
#> 26   -1.340323e+00      0
#> 27    2.983791e-01      0
#> 28    1.515901e+00      0
#> 29   -5.550184e-01      0
#> 30   -2.165068e+00      0
#> 31   -1.304005e-01      0
#> 32    3.557403e-01      0
#> 33    2.019351e-01      0
#> 34    5.310563e-01      0
#> 35    2.081197e+00      0
#> 36    5.320071e-01      0
#> 37   -5.490917e-01      0
#> 38   -3.434433e-01      0
#> 39    1.092172e+00      0
#> 40   -7.072781e-01      0
#> 41   -2.231916e-01      0
#> 42    9.504362e-01      0
#> 43   -3.638711e-01      0
#> 44   -5.413782e-01      0
#> 45   -5.775542e-01      0
#> 46   -3.383857e-01      0
#> 47    8.093359e-01      0
#> 48    6.396953e-01      0
#> 49    3.015925e-01      0
#> 50   -1.553893e+00      0
#> 51   -2.670399e-01      0
#> 52    8.052821e-01      0
#> 53    2.494143e-01      0
#> 54    1.655834e+00      0
#> 55    7.263269e-02      0
#> 56   -1.170128e-01      0
#> 57    9.054770e-01      0
#> 58   -5.092188e-01      0
#> 59   -1.973792e-01      0
#> 60   -1.525153e+00      0
#> 61   -5.759842e-01      0
#> 62    1.507550e+00      0
#> 63    6.115999e-01      0
#> 64   -2.106433e-01      0
#> 65   -3.187316e-01      0
#> 66   -9.502382e-01      0
#> 67   -6.401429e-02      0
#> 68    8.902395e-01      0
#> 69    8.339106e-01      0
#> 70    5.229098e-01      0
#> 71   -5.006133e-01      0
#> 72    9.198071e-02      0
#> 73   -1.726320e-01      0
#> 74    1.753249e+00      0
#> 75    6.204583e-01      0
#> 76   -7.868115e-01      0
#> 77   -4.987356e-01      0
#> 78    3.639444e-01      0
#> 79    8.525411e-01      0
#> 80   -5.529337e-01      0
#> 81   -6.604740e-01      0
#> 82   -2.169131e-02      0
#> 83   -8.335178e-01      0
#> 84    1.181800e+00      0
#> 85   -1.266593e+00      0
#> 86   -5.867760e-01      0
#> 87   -1.934303e+00      0
#> 88    7.722569e-01      0
#> 89   -4.928813e-01      0
#> 90    2.516956e-01      0
#> 91    2.185851e+00      0
#> 92   -6.012188e-01      0
#> 93    1.601065e+00      0
#> 94    6.752415e-02      0
#> 95    9.788640e-01      0
#> 96    1.308053e-01      0
#> 97    6.381302e-01      0
#> 98    6.536030e-01      0
#> 99   -1.449492e+00      0
#> 100   9.312133e-01      0
#> 101   5.503410e-01      0
#> 102   6.714486e-01      0
#> 103   3.917242e-01      0
#> 104  -1.457989e+00      0
#> 105   2.094354e-02      0
#> 106   8.223463e-01      0
#> 107  -6.646907e-01      0
#> 108  -4.615204e-01      0
#> 109   2.031109e+00      0
#> 110   3.977768e-01      0
#> 111  -3.690090e-01      0
#> 112   9.819883e-01      0
#> 113  -7.394628e-01      0
#> 114  -4.385871e-01      0
#> 115   7.906401e-01      0
#> 116  -1.583645e-01      0
#> 117   6.628954e-01      0
#> 118   6.384416e-01      0
#> 119   2.732434e-01      0
#> 120   8.979695e-01      0
#> 121   1.416974e+00      0
#> 122  -6.475399e-01      0
#> 123   5.091387e-01      0
#> 124  -1.086987e+00      0
#> 125  -1.014294e+00      0
#> 126  -4.429026e-01      0
#> 127  -9.868168e-01      0
#> 128  -4.833835e-01      0
#> 129  -1.059636e+00      0
#> 130   1.085315e+00      0
#> 131   2.959190e-01      0
#> 132   8.426835e-01      0
#> 133   1.618278e+00      0
#> 134  -3.221086e-01      0
#> 135  -2.055312e-01      0
#> 136  -9.247691e-01      0
#> 137   3.050720e+00      0
#> 138   3.634886e-01      0
#> 139   1.398248e+00      0
#> 140  -1.005175e-01      0
#> 141   3.939547e-01      0
#> 142   2.071278e-01      0
#> 143  -8.163702e-01      0
#> 144  -1.743673e+00      0
#> 145  -8.392761e-01      0
#> 146   3.528931e-02      0
#> 147  -9.728881e-01      0
#> 148  -4.396168e-01      0
#> 149  -4.221252e-01      0
#> 150  -1.020727e+00      0
#> 151  -3.227622e-01      0
#> 152   1.481348e+00      0
#> 153   5.212300e-01      0
#> 154  -9.318273e-01      0
#> 155  -3.332474e+00      0
#> 156   1.706776e+00      0
#> 157  -2.253870e-01      0
#> 158   1.267520e+00      0
#> 159   4.032016e-01      0
#> 160   3.267614e-02      0
#> 161  -1.561486e+00      0
#> 162   4.522934e-01      0
#> 163   1.330495e+00      0
#> 164  -2.417836e-01      0
#> 165  -3.630470e-01      0
#> 166  -4.616340e-01      0
#> 167   2.079026e-02      0
#> 168  -5.386048e-01      0
#> 169  -1.243632e-01      0
#> 170   1.352414e+00      0
#> 171  -4.482738e-01      0
#> 172   6.540035e-01      0
#> 173   3.067555e-01      0
#> 174   1.851645e+00      0
#> 175   1.262329e+00      0
#> 176   5.369341e-01      0
#> 177  -3.914048e-02      0
#> 178  -1.284573e+00      0
#> 179   5.084815e-02      0
#> 180   6.412515e-01      0
#> 181  -1.716537e-01      0
#> 182  -3.183715e-01      0
#> 183  -4.538993e-01      0
#> 184   2.369251e+00      0
#> 185  -8.148789e-01      0
#> 186  -1.147089e-01      0
#> 187   1.154125e-01      0
#> 188   1.016351e+00      0
#> 189   7.507177e-03      0
#> 190   6.867925e-01      0
#> 191  -5.212184e-02      0
#> 192   1.820569e+00      0
#> 193   5.461069e-01      0
#> 194   1.154172e+00      0
#> 195  -3.777842e-01      0
#> 196  -7.496822e-01      0
#> 197  -1.158904e+00      0
#> 198  -1.573393e-01      0
#> 199   3.375745e-01      0
#> 200   8.661457e-01      0
#> 201  -1.446243e+00      0
#> 202  -1.378176e+00      0
#> 203   4.808254e-01      0
#> 204  -1.945817e+00      0
#> 205  -1.309704e+00      0
#> 206  -5.782364e-01      0
#> 207   2.032751e+00      0
#> 208   1.281439e-01      0
#> 209   8.102800e-01      0
#> 210   4.804980e-01      0
#> 211  -1.701935e+00      0
#> 212  -3.518751e-01      0
#> 213  -8.428873e-01      0
#> 214   9.447219e-02      0
#> 215  -8.776362e-01      0
#> 216  -1.600745e+00      0
#> 217   1.283700e+00      0
#> 218   9.584904e-01      0
#> 219  -4.291354e-01      0
#> 220   4.785352e-01      0
#> 221  -2.576835e-01      0
#> 222  -2.281308e-01      0
#> 223   7.094695e-01      0
#> 224   8.592503e-01      0
#> 225   9.938219e-01      0
#> 226   1.148842e+00      0
#> 227  -2.409314e-01      0
#> 228  -1.700562e+00      0
#> 229   7.250669e-01      0
#> 230  -3.245684e-01      0
#> 231  -1.368249e-01      0
#> 232   4.911172e-01      0
#> 233  -1.556593e+00      0
#> 234  -6.277130e-01      0
#> 235  -7.137708e-02      0
#> 236  -1.484133e+00      0
#> 237  -1.332504e+00      0
#> 238  -1.351143e+00      0
#> 239  -1.822559e+00      0
#> 240  -4.403850e-01      0
#> 241  -8.312040e-01      0
#> 242   1.151052e-01      0
#> 243   8.054466e-01      0
#> 244  -8.612099e-01      0
#> 245   1.436564e-01      0
#> 246  -1.471009e+00      0
#> 247   2.018497e-01      0
#> 248  -1.592730e+00      0
#> 249   4.772355e-01      0
#> 250  -1.020497e+00      0
#> 251  -7.020218e-01      0
#> 252   1.399063e+00      0
#> 253  -5.637993e-01      0
#> 254  -1.891897e-02      0
#> 255   6.051506e-01      0
#> 256   1.357705e+00      0
#> 257   1.354420e+00      0
#> 258   5.730019e-01      0
#> 259   5.712523e-01      0
#> 260   1.991671e-01      0
#> 261  -1.432585e+00      0
#> 262  -1.668243e-01      0
#> 263  -9.362685e-01      0
#> 264   3.798734e-01      0
#> 265   3.813478e-01      0
#> 266  -8.456920e-01      0
#> 267   3.442376e-02      0
#> 268  -9.179383e-01      0
#> 269  -1.794532e+00      0
#> 270   9.689929e-01      0
#> 271  -4.375079e-01      0
#> 272   4.430924e-01      0
#> 273   1.898202e+00      0
#> 274  -2.672543e+00      0
#> 275  -5.984639e-01      0
#> 276   7.551604e-01      0
#> 277   6.879525e-01      0
#> 278  -1.334409e+00      0
#> 279   1.871055e-01      0
#> 280  -1.980232e+00      0
#> 281   2.226490e-01      0
#> 282   8.059649e-01      0
#> 283  -1.779791e+00      0
#> 284  -1.449458e+00      0
#> 285   2.430628e-01      0
#> 286   2.803584e-01      0
#> 287   2.005132e+00      0
#> 288   1.930618e+00      0
#> 289   4.446096e-02      0
#> 290   2.173433e+00      0
#> 291  -7.031397e-01      0
#> 292  -1.126996e-01      0
#> 293   4.583591e-01      0
#> 294  -1.486774e+00      0
#> 295   5.110477e-01      0
#> 296  -1.187045e+00      0
#> 297  -6.161163e-01      0
#> 298  -1.150872e+00      0
#> 299   9.748613e-01      0
#> 300   1.720536e+00      0
#> 301   4.632975e-01      0
#> 302  -6.474917e-01      0
#> 303   1.564116e+00      0
#> 304  -1.988903e+00      0
#> 305  -1.985603e+00      0
#> 306   1.021693e-01      0
#> 307  -3.481133e-01      0
#> 308  -1.162505e+00      0
#> 309  -1.830190e+00      0
#> 310  -6.577296e-02      0
#> 311  -7.661918e-01      0
#> 312  -2.983642e-01      0
#> 313   6.082611e-01      0
#> 314   7.548800e-01      0
#> 315   2.024556e-01      0
#> 316   1.053216e+00      0
#> 317  -1.254401e+00      0
#> 318   6.936902e-01      0
#> 319   3.409364e-01      0
#> 320  -1.149402e-02      0
#> 321   1.640393e+00      0
#> 322  -4.081243e-01      0
#> 323  -9.966753e-01      0
#> 324  -1.279628e+00      0
#> 325  -1.033395e+00      0
#> 326   1.956071e-01      0
#> 327   3.278222e-01      0
#> 328  -2.761011e+00      0
#> 329   4.770289e-01      0
#> 330   6.159648e-01      0
#> 331  -2.095971e-01      0
#> 332   5.467910e-02      0
#> 333  -4.677980e-01      0
#> 334  -3.838481e-01      0
#> 335   3.385501e-01      0
#> 336   1.411373e+00      0
#> 337  -1.752402e-01      0
#> 338  -1.029592e+00      0
#> 339  -3.201021e-01      0
#> 340  -3.548933e-01      0
#> 341  -2.555892e-01      0
#> 342  -4.346288e-01      0
#> 343   4.239625e-01      0
#> 344  -8.668269e-01      0
#> 345  -2.736194e+00      0
#> 346   6.301305e-01      0
#> 347  -7.316073e-02      0
#> 348  -2.893197e-02      0
#> 349  -4.523932e-01      0
#> 350   1.031936e+00      0
#> 351   4.330546e-01      0
#> 352  -1.468804e+00      0
#> 353   4.161404e-01      0
#> 354   1.747888e+00      0
#> 355   1.295530e+00      0
#> 356  -4.225900e-01      0
#> 357   1.550958e-01      0
#> 358  -1.074377e+00      0
#> 359   1.117826e+00      0
#> 360   9.280867e-01      0
#> 361   1.032799e+00      0
#> 362  -1.409592e+00      0
#> 363   3.069763e-01      0
#> 364  -7.301416e-01      0
#> 365   5.807717e-01      0
#> 366  -8.569506e-01      0
#> 367   1.005678e+00      0
#> 368  -1.071015e+00      0
#> 369  -1.265924e+00      0
#> 370  -2.033013e+00      0
#> 371  -1.367638e+00      0
#> 372  -1.196592e+00      0
#> 373  -1.357958e+00      0
#> 374  -1.415916e-02      0
#> 375   7.500882e-01      0
#> 376   1.181586e-01      0
#> 377  -1.397601e+00      0
#> 378   3.286411e-01      0
#> 379  -1.033668e+00      0
#> 380   3.853756e-01      0
#> 381  -2.327007e-01      0
#> 382   4.689415e-02      0
#> 383   7.075568e-01      0
#> 384   4.430254e-02      0
#> 385  -1.129491e+00      0
#> 386  -1.449973e+00      0
#> 387  -1.141421e+00      0
#> 388   1.156526e+00      0
#> 389   5.277279e-01      0
#> 390   2.638393e-01      0
#> 391  -4.265547e-01      0
#> 392   6.661354e-01      0
#> 393   6.191520e-01      0
#> 394   1.622474e+00      0
#> 395  -1.092633e+00      0
#> 396   5.181139e-01      0
#> 397   7.284125e-01      0
#> 398   1.446251e+00      0
#> 399  -3.211288e-01      0
#> 400  -8.161924e-01      0
#> 401   1.240232e+00      0
#> 402   2.935999e-01      0
#> 403   1.187864e+00      0
#> 404   5.170517e-01      0
#> 405  -1.232987e+00      0
#> 406  -5.023528e-01      0
#> 407  -2.185466e-01      0
#> 408  -1.631332e+00      0
#> 409   1.333578e+00      0
#> 410   1.770187e+00      0
#> 411  -1.155586e+00      0
#> 412   4.017885e-01      0
#> 413   1.729592e+00      0
#> 414  -4.445292e-01      0
#> 415   4.943686e-01      0
#> 416   1.709263e+00      0
#> 417  -5.848567e-01      0
#> 418  -1.100505e-01      0
#> 419   8.812815e-01      0
#> 420  -3.364807e-01      0
#> 421   3.618738e-01      0
#> 422  -1.949965e+00      0
#> 423   8.922379e-01      0
#> 424   8.066956e-01      0
#> 425   6.335990e-01      0
#> 426   5.151167e-01      0
#> 427   2.884578e-01      0
#> 428  -5.391372e-01      0
#> 429  -2.147504e-01      0
#> 430  -3.275875e-01      0
#> 431   1.152111e+00      0
#> 432  -3.741767e-01      0
#> 433  -1.392470e+00      0
#> 434  -2.017578e-01      0
#> 435   9.585034e-02      0
#> 436   6.986647e-01      0
#> 437   1.266704e-01      0
#> 438  -4.809899e-01      0
#> 439  -2.287989e+00      0
#> 440   1.518810e-01      0
#> 441  -1.528603e+00      0
#> 442   9.656443e-01      0
#> 443   5.332841e-01      0
#> 444   3.015168e-01      0
#> 445  -1.744084e+00      0
#> 446   7.650911e-02      0
#> 447   1.907564e+00      0
#> 448   4.245545e-01      0
#> 449   1.076713e+00      0
#> 450   4.065173e-01      0
#> 451   1.065344e+00      0
#> 452  -8.691003e-01      0
#> 453  -2.111501e-02      0
#> 454  -6.990064e-01      0
#> 455  -1.012321e+00      0
#> 456  -5.455669e-01      0
#> 457   6.729648e-01      0
#> 458  -1.797474e+00      0
#> 459  -3.141021e-01      0
#> 460   1.353756e+00      0
#> 461  -7.264353e-01      0
#> 462   2.860028e-01      0
#> 463  -7.809137e-01      0
#> 464   4.489313e-01      0
#> 465   1.298943e+00      0
#> 466  -1.516862e+00      0
#> 467   2.218568e+00      0
#> 468  -2.809671e-01      0
#> 469  -5.559522e-01      0
#> 470   1.399661e-01      0
#> 471   5.242856e-02      0
#> 472  -1.985011e-01      0
#> 473  -4.284429e-01      0
#> 474  -4.323650e-01      0
#> 475  -3.969799e-03      0
#> 476   1.130156e+00      0
#> 477  -1.161683e-01      0
#> 478   8.875001e-01      0
#> 479  -6.950519e-01      0
#> 480  -8.852778e-01      0
#> 481  -3.167036e-01      0
#> 482   4.337358e-01      0
#> 483   1.374253e-01      0
#> 484  -5.657858e-01      0
#> 485   2.044767e-01      0
#> 486   8.824286e-01      0
#> 487   1.175269e+00      0
#> 488   3.664386e-01      0
#> 489  -4.221234e-01      0
#> 490   1.492116e+00      0
#> 491  -1.777908e+00      0
#> 492   8.712919e-01      0
#> 493   1.230901e+00      0
#> 494   1.113349e+00      0
#> 495   6.732128e-01      0
#> 496   4.835809e-01      0
#> 497  -2.007206e-02      0
#> 498   4.326202e-01      0
#> 499  -7.344549e-01      0
#> 500  -4.938705e-01      0
#> 501   1.473288e+00      0
#> 502   5.013794e-01      0
#> 503   1.655680e-01      0
#> 504   2.344694e-01      0
#> 505   1.928685e+00      0
#> 506  -2.431708e-02      0
#> 507   2.170265e-01      0
#> 508   1.124584e+00      0
#> 509   1.114067e-01      0
#> 510  -1.097568e+00      0
#> 511   4.987684e-01      0
#> 512  -4.525989e-01      0
#> 513  -2.234193e-01      0
#> 514   9.427766e-01      0
#> 515  -2.239643e+00      0
#> 516  -3.354571e-01      0
#> 517   9.208822e-01      0
#> 518   3.503578e-01      0
#> 519  -1.205968e+00      0
#> 520  -1.422772e+00      0
#> 521  -1.368506e+00      0
#> 522   6.211548e-01      0
#> 523   3.203375e-01      0
#> 524   5.308016e-01      0
#> 525   1.910809e-01      0
#> 526  -1.175880e+00      0
#> 527  -3.405210e-01      0
#> 528   1.468214e+00      0
#> 529   1.905963e-01      0
#> 530  -1.806237e+00      0
#> 531   1.660732e+00      0
#> 532   1.591797e+00      0
#> 533  -4.033618e-01      0
#> 534  -4.262429e-01      0
#> 535   9.346246e-01      0
#> 536  -3.480407e-01      0
#> 537   3.344601e-01      0
#> 538  -2.253701e+00      0
#> 539   3.687035e-01      0
#> 540   4.421562e-01      0
#> 541  -2.475377e+00      0
#> 542  -5.649483e-01      0
#> 543   1.430995e-01      0
#> 544   8.483440e-01      0
#> 545  -3.623693e-01      0
#> 546  -5.742117e-01      0
#> 547  -6.095912e-03      0
#> 548   4.818662e-01      0
#> 549  -9.007446e-02      0
#> 550  -5.633781e-01      0
#> 551   6.182076e-02      0
#> 552  -1.728771e+00      0
#> 553  -1.555725e+00      0
#> 554  -2.142000e+00      0
#> 555   1.329228e+00      0
#> 556   1.780578e-01      0
#> 557  -3.025640e-01      0
#> 558   5.349963e-02      0
#> 559   4.102882e-01      0
#> 560  -2.776375e-03      0
#> 561   1.869900e+00      0
#> 562  -7.121310e-01      0
#> 563  -9.868048e-01      0
#> 564  -1.053728e-02      0
#> 565  -1.356476e-02      0
#> 566   5.330178e-01      0
#> 567  -6.972912e-01      0
#> 568   4.247276e-01      0
#> 569   1.320834e-01      0
#> 570   1.166353e+00      0
#> 571   7.758902e-01      0
#> 572  -1.383156e+00      0
#> 573  -1.846026e+00      0
#> 574   8.497611e-02      0
#> 575   7.806489e-01      0
#> 576  -8.103020e-02      0
#> 577   1.162308e+00      0
#> 578   1.157878e-01      0
#> 579  -6.248638e-01      0
#> 580  -6.814331e-01      0
#> 581  -7.301275e-01      0
#> 582  -9.252508e-01      0
#> 583   2.640607e-02      0
#> 584   3.416568e-01      0
#> 585   6.250249e-01      0
#> 586  -9.441043e-01      0
#> 587  -6.368514e-01      0
#> 588  -7.346699e-01      0
#> 589  -3.636449e-01      0
#> 590  -1.133799e+00      0
#> 591  -1.800354e-01      0
#> 592  -1.217294e-01      0
#> 593  -8.829801e-01      0
#> 594  -1.018036e+00      0
#> 595  -1.489846e+00      0
#> 596  -1.107352e-01      0
#> 597   3.513596e-01      0
#> 598   7.951028e-01      0
#> 599   9.376427e-02      0
#> 600   5.710562e-01      0
#> 601   4.448122e-01      0
#> 602  -7.127088e-01      0
#> 603   2.258613e-01      0
#> 604   3.452010e-01      0
#> 605   6.728930e-01      0
#> 606   1.065196e+00      0
#> 607  -9.448634e-01      0
#> 608   1.084786e-01      0
#> 609   9.428350e-01      0
#> 610  -1.032337e+00      0
#> 611  -5.770504e-02      0
#> 612   2.254150e-01      0
#> 613  -9.641127e-01      0
#> 614   4.968944e-01      0
#> 615   1.630260e+00      0
#> 616   1.892218e+00      0
#> 617  -4.909231e-01      0
#> 618  -9.787513e-01      0
#> 619  -1.439664e+00      0
#> 620   4.563890e-01      0
#> 621  -1.335117e-01      0
#> 622  -1.425431e+00      0
#> 623   2.342291e+00      0
#> 624  -1.687092e-01      0
#> 625   3.077057e-01      0
#> 626   6.049314e-01      0
#> 627  -8.814373e-01      0
#> 628  -2.531818e+00      0
#> 629  -2.098363e+00      0
#> 630  -4.445291e-01      0
#> 631   1.026421e+00      0
#> 632  -8.382889e-01      0
#> 633  -1.280051e+00      0
#> 634  -1.221961e+00      0
#> 635  -8.005997e-01      0
#> 636   7.452557e-02      0
#> 637  -1.065966e+00      0
#> 638  -1.829303e+00      0
#> 639   7.077207e-01      0
#> 640  -1.232135e+00      0
#> 641   6.549067e-01      0
#> 642   2.062471e+00      0
#> 643   4.145657e-02      0
#> 644  -1.013400e+00      0
#> 645   4.193986e-01      0
#> 646  -6.866168e-01      0
#> 647  -6.706073e-01      0
#> 648   1.996176e+00      0
#> 649  -7.935663e-02      0
#> 650   1.688887e+00      0
#> 651   2.176439e+00      0
#> 652  -1.105143e+00      0
#> 653   1.058556e+00      0
#> 654   2.696349e-01      0
#> 655   9.659015e-01      0
#> 656   6.982914e-01      0
#> 657  -6.682815e-01      0
#> 658   5.141099e-01      0
#> 659  -6.481946e-01      0
#> 660   7.885634e-01      0
#> 661   1.482547e+00      0
#> 662   3.009911e+00      0
#> 663   1.040902e+00      0
#> 664   9.211232e-01      0
#> 665   4.492339e-01      0
#> 666   5.554906e-01      0
#> 667  -2.070516e+00      0
#> 668  -1.251068e+00      0
#> 669   1.084618e+00      0
#> 670  -1.560763e+00      0
#> 671  -1.224107e+00      0
#> 672  -1.645235e+00      0
#> 673   6.254698e-01      0
#> 674   7.855585e-01      0
#> 675   7.754257e-02      0
#> 676   7.964002e-01      0
#> 677  -3.986238e-01      0
#> 678   2.367380e-01      0
#> 679  -4.894784e-01      0
#> 680  -3.786385e-01      0
#> 681  -1.095270e+00      0
#> 682   5.452189e-01      0
#> 683   5.038713e-01      0
#> 684  -2.006250e+00      0
#> 685   1.452299e+00      0
#> 686  -4.139027e-01      0
#> 687  -6.611718e-01      0
#> 688  -1.126280e-01      0
#> 689  -2.376561e-01      0
#> 690  -3.377109e-01      0
#> 691  -3.513653e-01      0
#> 692  -3.077632e-01      0
#> 693  -1.609440e+00      0
#> 694   3.229608e-01      0
#> 695   1.114668e+00      0
#> 696   3.867013e-01      0
#> 697  -2.079729e-01      0
#> 698  -1.002371e+00      0
#> 699  -4.468655e-01      0
#> 700  -3.053567e-01      0
#> 701  -1.001141e+00      0
#> 702  -6.848845e-05      0
#> 703  -7.873091e-02      0
#> 704  -1.090887e-01      0
#> 705   7.871988e-01      0
#> 706   5.105160e-01      0
#> 707   2.473227e+00      0
#> 708   1.523073e+00      0
#> 709  -1.081599e+00      0
#> 710   1.248401e+00      0
#> 711   1.283306e+00      0
#> 712  -5.415388e-01      0
#> 713  -2.525720e-01      0
#> 714   1.154520e+00      0
#> 715   1.603678e+00      0
#> 716   3.223746e-01      0
#> 717  -1.241859e+00      0
#> 718  -1.133160e-01      0
#> 719   5.249237e-01      0
#> 720  -8.312361e-01      0
#> 721  -3.511264e-01      0
#> 722   2.117127e-01      0
#> 723   1.609299e+00      0
#> 724  -2.818072e-01      0
#> 725   8.968364e-02      0
#> 726  -1.515060e-01      0
#> 727   1.157112e+00      0
#> 728  -5.579369e-01      0
#> 729   3.394332e-01      0
#> 730   1.297190e+00      0
#> 731   1.556328e-02      0
#> 732   4.053092e-01      0
#> 733  -7.601209e-01      0
#> 734   2.424381e+00      0
#> 735   5.903601e-01      0
#> 736  -7.421162e-01      0
#> 737  -9.537172e-01      0
#> 738  -1.502466e+00      0
#> 739  -2.804314e-01      0
#> 740  -1.090041e-02      0
#> 741   4.690422e-01      0
#> 742   1.221215e+00      0
#> 743  -3.667558e-01      0
#> 744  -6.644267e-01      0
#> 745   9.144445e-01      0
#> 746   7.655134e-01      0
#> 747   9.036323e-01      0
#> 748  -6.535154e-01      0
#> 749  -5.396148e-01      0
#> 750   9.829632e-01      0
#> 751  -2.817671e-02      0
#> 752   1.372754e+00      0
#> 753  -7.546752e-01      0
#> 754  -7.949527e-01      0
#> 755   1.724758e-01      0
#> 756   1.554772e+00      0
#> 757   5.168359e-01      0
#> 758   6.994467e-01      0
#> 759  -6.422057e-01      0
#> 760   5.220140e-01      0
#> 761  -1.870293e+00      0
#> 762  -5.719924e-01      0
#> 763  -9.062102e-01      0
#> 764   3.834348e-01      0
#> 765   6.730821e-02      0
#> 766  -5.113415e-01      0
#> 767   1.378389e+00      0
#> 768  -5.832244e-01      0
#> 769   2.232137e+00      0
#> 770   1.583287e+00      0
#> 771  -7.866550e-01      0
#> 772   1.414500e+00      0
#> 773  -6.662045e-02      0
#> 774   1.442451e+00      0
#> 775   3.173192e-02      0
#> 776  -8.798932e-01      0
#> 777   4.077089e-01      0
#> 778   9.983721e-01      0
#> 779   4.997902e-01      0
#> 780   1.181900e-01      0
#> 781   2.758790e-01      0
#> 782  -7.147290e-01      0
#> 783   1.046186e+00      0
#> 784  -2.041043e-01      0
#> 785   3.947663e-01      0
#> 786   1.413806e+00      0
#> 787  -3.793951e-01      0
#> 788  -5.789462e-01      0
#> 789   2.420802e-01      0
#> 790   8.388023e-01      0
#> 791  -6.704129e-01      0
#> 792  -5.440450e-01      0
#> 793   1.457589e+00      0
#> 794  -1.079746e+00      0
#> 795   2.891607e-01      0
#> 796  -7.187001e-01      0
#> 797  -9.433291e-01      0
#> 798  -4.300727e-01      0
#> 799   1.362510e+00      0
#> 800  -1.774144e-01      0
#> 801   1.331783e-01      0
#> 802   1.240458e+00      0
#> 803   4.328906e-01      0
#> 804   2.075732e+00      0
#> 805   3.062018e-02      0
#> 806   3.995769e-01      0
#> 807   7.876854e-01      0
#> 808   3.553047e-01      0
#> 809  -3.799134e-02      0
#> 810   1.445689e+00      0
#> 811   1.098456e+00      0
#> 812   3.502327e-02      0
#> 813   1.163833e+00      0
#> 814  -1.623724e-03      0
#> 815   1.680875e-01      0
#> 816  -4.035336e-01      0
#> 817  -1.523539e+00      0
#> 818   1.011464e+00      0
#> 819   2.824991e-01      0
#> 820  -6.328402e-01      0
#> 821  -1.856413e-01      0
#> 822   4.856533e-01      0
#> 823   5.215565e-01      0
#> 824  -1.022175e+00      0
#> 825  -1.507946e+00      0
#> 826  -1.041049e+00      0
#> 827   1.099128e+00      0
#> 828   3.029527e-01      0
#> 829   1.473523e+00      0
#> 830  -2.595688e-01      0
#> 831   1.330511e+00      0
#> 832   1.000028e-02      0
#> 833  -9.740424e-01      0
#> 834   2.143890e-01      0
#> 835  -6.016502e-03      0
#> 836   1.214605e+00      0
#> 837   8.178373e-01      0
#> 838   1.640455e+00      0
#> 839   1.511473e+00      0
#> 840   4.410278e-01      0
#> 841   1.443254e+00      0
#> 842  -2.222426e-01      0
#> 843  -6.750053e-02      0
#> 844   1.152810e+00      0
#> 845   7.740704e-01      0
#> 846   4.632089e-01      0
#> 847   1.539061e-02      0
#> 848  -1.448325e+00      0
#> 849   8.677008e-01      0
#> 850  -5.623029e-01      0
#> 851   1.119453e+00      0
#> 852   1.279739e+00      0
#> 853   5.777272e-01      0
#> 854  -6.590730e-02      0
#> 855  -1.059387e+00      0
#> 856  -1.253482e+00      0
#> 857  -3.028182e-01      0
#> 858   1.096089e+00      0
#> 859  -9.085908e-01      0
#> 860   8.608077e-01      0
#> 861  -2.375735e+00      0
#> 862  -3.349588e-01      0
#> 863   1.354439e+00      0
#> 864  -1.379051e-01      0
#> 865  -1.528351e+00      0
#> 866   7.216904e-01      0
#> 867  -2.383446e-01      0
#> 868   5.344338e-01      0
#> 869  -6.860474e-02      0
#> 870   9.852906e-01      0
#> 871  -2.100202e-01      0
#> 872  -6.381753e-01      0
#> 873   1.246357e+00      0
#> 874  -9.038390e-02      0
#> 875   1.525974e-01      0
#> 876   1.471779e+00      0
#> 877   4.803652e-01      0
#> 878  -9.409851e-01      0
#> 879  -8.061849e-02      0
#> 880   1.723557e+00      0
#> 881   8.449999e-01      0
#> 882   4.904235e-01      0
#> 883   1.423671e+00      0
#> 884   2.248131e-01      0
#> 885  -4.619529e-01      0
#> 886  -1.408304e+00      0
#> 887  -4.854995e-01      0
#> 888  -4.369537e-01      0
#> 889   9.148215e-01      0
#> 890   9.208609e-01      0
#> 891   6.366303e-01      1
#> 892   2.080761e+00      1
#> 893   1.213466e+00      1
#> 894   9.337662e-01      1
#> 895  -3.662743e-01      1
#> 896   2.171133e+00      1
#> 897   2.841840e+00      1
#> 898   2.218653e+00      1
#> 899   7.845127e-01      1
#> 900   8.537105e-02      1
#> 901   8.084320e-02      1
#> 902   1.089807e+00      1
#> 903   1.506507e-01      1
#> 904   2.989693e-01      1
#> 905   1.167690e+00      1
#> 906   2.441282e-01      1
#> 907   9.298280e-01      1
#> 908  -5.866845e-01      1
#> 909   2.064751e+00      1
#> 910   1.747045e+00      1
#> 911   1.771436e+00      1
#> 912   2.679947e+00      1
#> 913   2.647902e+00      1
#> 914   8.730742e-01      1
#> 915   1.189019e+00      1
#> 916   1.548395e-01      1
#> 917  -3.911551e-01      1
#> 918  -1.323290e+00      1
#> 919   2.272729e-01      1
#> 920   4.203234e-01      1
#> 921   1.555956e+00      1
#> 922   4.144768e-01      1
#> 923   2.863261e+00      1
#> 924   2.447231e-01      1
#> 925  -1.138904e-01      1
#> 926   1.164352e+00      1
#> 927   7.028803e-01      1
#> 928  -5.035402e-01      1
#> 929   2.460447e+00      1
#> 930   8.416362e-01      1
#> 931   8.768678e-01      1
#> 932   1.352943e+00      1
#> 933   2.138510e+00      1
#> 934   1.151938e+00      1
#> 935   2.211048e-01      1
#> 936   1.437401e-01      1
#> 937   6.373764e-01      1
#> 938   7.494231e-01      1
#> 939   8.678254e-01      1
#> 940   2.713240e+00      1
#> 941   2.330311e-01      1
#> 942   5.423343e-01      1
#> 943   1.281706e+00      1
#> 944   8.487237e-01      1
#> 945   1.073506e+00      1
#> 946  -1.574679e-01      1
#> 947  -3.819744e-01      1
#> 948   6.448970e-01      1
#> 949   1.974869e+00      1
#> 950   3.936758e-01      1
#> 951   2.272090e+00      1
#> 952   8.449818e-01      1
#> 953   6.250299e-01      1
#> 954   1.067821e+00      1
#> 955   1.575964e+00      1
#> 956  -5.857074e-01      1
#> 957   1.010110e+00      1
#> 958   1.203290e+00      1
#> 959  -5.497406e-01      1
#> 960   3.278818e-01      1
#> 961  -4.214360e-01      1
#> 962   9.033558e-01      1
#> 963   1.997867e-01      1
#> 964   3.847782e-01      1
#> 965   1.413271e+00      1
#> 966   4.920858e-02      1
#> 967   1.196388e+00      1
#> 968  -1.462428e-01      1
#> 969   1.419760e-01      1
#> 970  -8.473845e-02      1
#> 971   8.485333e-01      1
#> 972  -7.126951e-01      1
#> 973   4.264944e-01      1
#> 974  -5.501710e-01      1
#> 975   1.505277e+00      1
#> 976   6.194799e-01      1
#> 977   9.269943e-01      1
#> 978   2.510116e-01      1
#> 979   1.499670e-01      1
#> 980  -4.073558e-01      1
#> 981   2.422143e+00      1
#> 982   4.131022e-01      1
#> 983  -5.974390e-01      1
#> 984   1.840685e+00      1
#> 985  -1.364539e+00      1
#> 986   4.700771e-01      1
#> 987   1.083301e+00      1
#> 988  -9.659331e-03      1
#> 989   1.154348e+00      1
#> 990   1.582879e+00      1
#> 991   2.141386e+00      1
#> 992   1.125594e+00      1
#> 993  -8.707637e-01      1
#> 994   7.185239e-01      1
#> 995   8.226715e-01      1
#> 996   4.499251e-01      1
#> 997   8.532961e-01      1
#> 998  -8.155367e-01      1
#> 999   5.055667e-01      1
#> 1000  2.369095e+00      1