Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.8854244        0.7027400 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4480811        0.4480811        0.4716824 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.257660643430427, 0.283676178893074, 0.0140263086650521, 0.417393579846248, 0.175468749832362, 0.359796206466854, 0.43725808034651, 0.0969232378993183, 0.863016183953732, 0.272942404262722, 0.339907473884523, 0.647286639548838, 0.743045416194946, 0.335894226562232, 0.347321576904505, 0.733688906533644, 0.804624191718176, 0.155840539606288, 0.265231683617458, 0.00447143893688917, 0.936523141572252, 0.25065566925332, 0.265302315587178, 0.507636315422133, 0.0528019035700709, 0.459405326517299, 0.838720151456073, 0.280178716406226, 0.865497739519924, 0.511512502096593, 0.468139347620308, 0.856346837477759, 0.237747110892087, 0.406170037807897, 0.272569192107767, 0.374559306073934, 0.746268094051629, 0.676114576635882, 0.227070036577061, 0.00997234485112131, 0.396917945472524, 0.931564143626019, 0.798577995738015, 0.923954564612359, 0.474183975020424, 0.826495066750795, 0.481665096245706, 0.463196200318635, 0.927244164980948, 0.74494355591014, 0.874977828003466, 0.654131461633369, 0.840528276748955, 0.304669614881277, 0.126176561228931, 0.952136107254773, 0.322059305850416, 0.28967038472183, 0.0897683687508106, 0.285793081391603, 0.261242919368669, 0.974155193893239, 0.377332261530682, 0.853098345687613, 0.749955820851028, 0.221538174664602, 0.368859060341492, 0.403368955943733, 0.464383908780292, 0.769328600494191, 0.824326494010165, 0.612405988853425, 0.788067637244239, 0.764607620891184, 0.475199478212744, 0.449683702550828, 0.455772146582603, 0.524206763366237, 0.456709810998291, 0.567511645378545, 0.919122139457613, 0.170384896919131, 0.804842810612172, 0.366869118763134, 0.0555930046830326, 0.989037954481319, 0.420968062942848, 0.533487889450043, 0.9844813386444, 0.40740384045057, 0.345632976852357, 0.0113176465965807, 0.71442165132612, 0.71688744192943, 0.69912021397613, 0.909942254191265, 0.933880841825157, 0.566971110878512, 0.740858980920166, 0.715365293901414, 0.484863507794216, 0.946214808383957, 0.853049710160121, 0.518001401796937, 0.732087171170861, 0.749589198501781, 0.627362021710724, 0.673283973010257, 0.85761228739284, 0.226758426520973, 0.119985034223646, 0.0752892326563597, 0.400481923250481, 0.172674940200523, 0.533507315209135, 0.670094284694642, 0.204572970513254, 0.73970390134491, 0.0219163442961872, 0.0512195287737995, 0.73677781643346, 0.125595025252551, 0.0951648645568639, 0.507136628963053, 0.879693920025602, 0.578769655432552, 0.616420242236927, 0.302381162066013, 0.894910129485652, 0.712962265824899, 0.780685001518577, 0.28899449063465, 0.4146212216001, 0.97035407088697, 0.778567427536473, 0.70677336701192, 0.136374476831406, 0.638435355387628, 0.207349261268973, 0.0333626000210643, 0.577587749110535, 0.180289119947702, 0.33357675280422, 0.513358000665903, 0.272196436766535, 0.379979741293937, 0.48413454182446, 0.459038197994232, 0.627792821964249, 0.693651112029329, 0.216161911608651, 0.894080623751506, 0.0450951769016683, 0.980571202468127, 0.863316820235923, 0.371719415998086, 0.223853333853185, 0.199772570515051, 0.541896526003256, 0.128798942547292, 0.34471838362515, 0.205859476467595, 0.766954925144091, 0.503912307787687, 0.0693509646225721, 0.385759299155325, 0.378822484053671, 0.546424374449998, 0.963181101484224, 0.984185602748767, 0.115917623275891, 0.845671562710777, 0.186680623330176, 0.586659162770957, 0.0182842186186463, 0.895332723390311, 0.890590694965795, 0.35754374274984, 0.278522222070023, 0.0149142893496901, 0.41858249460347, 0.493299916852266, 0.627786880359054, 0.559933594660833, 0.0509557237382978, 0.35577572346665, 0.530895517440513, 0.74930456536822, 0.560179888037965, 0.250498824287206, 0.908700661500916, 0.546420267783105, 0.28334262012504, 0.0187290858011693, 0.275037416722625, 0.174751647748053, 0.471910149324685, 0.510157039621845, 0.821989410324022, 0.0712511902675033, 0.315930784679949, 0.599052544683218, 0.750289633637294, 0.368112954776734, 0.515123199205846, 0.311127793509513, 0.792306142160669, 0.444732333999127, 0.728577963775024, 0.824849329190329, 0.278300474863499, 0.263373586814851, 0.0705024790950119, 0.751944929128513, 0.72776466421783, 0.397620627656579, 0.0740801107604057, 0.228269173996523, 0.0944081693887711, 0.295298758661374, 0.526518631028011, 0.262194897048175, 0.957048415439203, 0.786996577400714, 0.64545095525682, 0.292625203728676, 0.771995690185577, 0.997506494401023, 0.782017888734117, 0.58156535984017, 0.879327555885538, 0.461100246524438, 0.194991296390072, 0.00858483952470124, 0.661739789880812, 0.398343367036432, 0.789686844684184, 0.885653056204319, 0.633420441998169, 0.607782175997272, 0.0892714974470437, 0.9255871551577, 0.78433374222368, 0.161653423914686, 0.830048870760947, 0.164292334811762, 0.673013959545642, 0.32477887510322, 0.981109202140942, 0.664694169303402, 0.153137909015641, 0.975189189892262, 0.477362493285909, 0.0395196639001369, 0.910359058761969, 0.663350760238245, 0.282358530675992, 0.95503193908371, 0.129263529554009, 0.0423426004126668, 0.178744968958199, 0.123109050327912, 0.535684172762558, 0.956874579656869, 0.223141046473756, 0.590712526813149, 0.37113973335363, 0.417416249867529, 0.587758410489187, 0.0421077264472842, 0.17177672823891, 0.700790995499119, 0.160138288512826, 0.0696893080603331, 0.34906815038994, 0.00867382041178644, 0.0856240012217313, 0.86200263723731, 0.190799606498331, 0.759339454118162, 0.490961266215891, 0.524936826899648, 0.729882192797959, 0.193728194339201, 0.212872748961672, 0.282711022300646, 0.091768441721797, 0.309522861847654, 0.417879438726231, 0.793788542738184, 0.134420041227713, 0.723407172365114, 0.172432287130505, 0.87745984015055, 0.911807124502957, 0.566080002347007, 0.645906718447804, 0.532162262126803, 0.423748773289844, 0.00392660382203758, 0.0406241423916072, 0.323651894927025, 0.235064744250849, 0.985838883556426, 0.475160799920559, 0.941812518984079, 0.349631374236196, 0.0690788885112852, 0.163486274890602, 0.176693863002583, 0.0635698714759201, 0.693283544387668, 0.284658331191167, 0.194055500905961, 0.526481410721317, 0.814809748204425, 0.844228389207274, 0.710940026445314, 0.891395696438849, 0.998409893596545, 0.404052045661956, 0.328128659399226, 0.591530825477093, 0.0352349709719419, 0.507522815139964, 0.959109418326989, 0.413024072768167, 0.489864534931257, 0.269100345904008, 0.728972719283774, 0.177917960798368, 0.595984175102785, 0.136304183397442, 0.780860403552651, 0.420134127838537, 0.831681165611371, 0.951736627845094, 0.94854955538176, 0.7009093365632, 0.48300104169175, 0.690194342983887, 0.514600295107812, 0.853436522418633, 0.791573375463486, 0.469024929683656, 0.210083471378312, 0.223476310260594, 0.702848788583651, 0.967179729603231, 0.918014761526138, 0.911016863305122, 0.345093386946246, 0.167819241993129, 0.863013480324298, 0.610901717329398, 0.928974475944415, 0.0904805564787239, 0.4253903364297, 0.195091880625114, 0.632716292282566, 0.471368500497192, 0.305837720399722, 0.0282938757445663, 0.509280447149649, 0.63898440846242, 0.20998139702715, 0.0579604371450841, 0.573973515536636, 0.600516442907974, 0.775069387629628, 0.508793527958915, 0.436104700900614, 0.0748600091319531, 0.441982357529923, 0.969198386417702, 0.166675666347146, 0.0676635976415128, 0.466335866600275, 0.608892053365707, 0.75924199540168, 0.459504744736478, 0.473620641045272, 0.153097131988034, 0.836040051421151, 0.681879320414737, 0.0464154167566448, 0.45272845774889, 0.0684642414562404, 0.75860058539547, 0.956063566729426, 0.670965891331434, 0.11141765746288, 0.999887107405812, 0.299879176542163, 0.57991756615229, 0.111686605494469, 0.59214151930064, 0.331999726360664, 0.0923490424174815, 0.0677122317720205, 0.247505732346326, 0.911561302142218, 0.688834508880973, 0.608602575724944, 0.318446569610387, 0.841380966827273, 0.561886393697932, 0.398419623961672, 0.77773735509254, 0.941845038440078, 0.838404988870025, 0.0468711976427585, 0.0611224635504186, 0.542085461085662, 0.626070421887562, 0.952664729440585, 0.0510880453512073, 0.453706218861043, 0.305816289270297, 0.220595143502578, 0.735036299098283, 0.644174759043381, 0.590728212846443, 0.645035687834024, 0.406401766929775, 0.714454696746543, 0.402045347727835, 0.495143041014671, 0.0489705004729331, 0.829555869335309, 0.859316414222121, 0.622377233579755, 0.623090036446229, 0.529328807257116, 0.117229925235733, 0.319576033623889, 0.617894791299477, 0.573635149979964, 0.900240666233003, 0.172127592144534, 0.853434630902484, 0.858869303250685, 0.845727710751817, 0.141525669256225, 0.664949883474037, 0.442912962287664, 0.274156822590157, 0.80664837337099, 0.70341779384762, 0.219922966323793, 0.208228460280225, 0.379169685766101, 0.717282605590299, 0.914198711514473, 0.706673628650606, 0.152624465059489, 0.464150127954781, 0.493005607044324, 0.48851968231611, 0.159897328121588, 0.862966337008402, 0.60410240595229, 0.189184207934886, 0.35010218876414, 0.190105990506709, 0.0470570230390877, 0.609241920057684, 0.0865975259803236, 0.494455559877679, 0.962944756262004, 0.91702613607049, 0.742684702388942, 0.657990620005876, 0.772717723622918, 0.950311791384593, 0.261223967652768, 0.104637879645452, 0.857443284010515, 0.345793198561296, 0.647841995116323, 0.625068813096732, 0.60937225818634, 0.523691104026511, 0.50602908199653, 0.39677740004845, 0.269614382414147, 0.456018171273172, 0.756330548319966, 0.0495800510980189, 0.479674263857305, 0.935834704898298, 0.302043337374926, 0.23087988095358, 0.371269157389179, 0.751774365780875, 0.938115192577243, 0.453591150231659, 0.207248501712456, 0.304412094177678, 0.553548890864477, 0.68247422366403, 0.288428530097008, 0.696665436727926, 0.737307009054348, 0.921600674279034, 0.549112578155473, 0.873612158000469, 0.0110841512214392, 0.593453987734392, 0.701431603636593, 0.678282868582755, 0.467217304045334, 0.0267025462817401, 0.124401243403554, 0.412118094973266, 0.985767158446833, 0.100034457165748, 0.0115678852889687, 0.273726407904178, 0.935327455634251, 0.477596591226757, 0.658909102668986, 0.741572776809335, 0.961833368521184, 0.638097636634484, 0.308282893849537, 0.782673930749297, 0.430305574322119, 0.76035764766857, 0.613630110397935, 0.129359256941825, 0.833225244423375, 0.783857703907415, 0.53641292732209, 0.305342808831483, 0.845071319257841, 0.993379798252136, 0.949807926546782, 0.768250452820212, 0.746354257687926, 0.810116331558675, 0.27514585480094, 0.249255133327097, 0.611407853895798, 0.767333635361865, 0.305672206450254, 0.907680805306882, 0.540010415483266, 0.418515038210899, 0.921033482765779, 0.942300033988431, 0.196493654511869, 0.78468731045723, 0.869628384476528, 0.0801129464525729, 0.232352801365778, 0.444727249443531, 0.625797646353021, 0.513624306302518, 0.109844112768769, 0.785775488475338, 0.694631838705391, 0.903365618549287, 0.589288971619681, 0.411361539503559, 0.296820324612781, 0.0246397312730551, 0.998498706845567, 0.408931475132704, 0.966972271446139, 0.132203077897429, 0.988012094749138, 0.809450334636495, 0.781608372461051, 0.535212934948504, 0.81251356122084, 0.906757676973939, 0.111200981540605, 0.729544303612784, 0.229738882277161, 0.163526856573299, 0.694709893548861, 0.0819799120072275, 0.309760480653495, 0.0541278310120106, 0.273819925729185, 0.890550228301436, 0.880237625679001, 0.68834074260667, 0.851080352906138, 0.631686181994155, 0.964514565421268, 0.171015918022022, 0.542233735555783, 0.784610135946423, 0.745148953283206, 0.837855949997902, 0.361638487316668, 0.595044837566093, 0.434400361031294, 0.0917341040913016, 0.126806110842153, 0.0777849885635078, 0.238397912820801, 0.659010056173429, 0.543179427972063, 0.223328829975799, 0.103878219844773, 0.754103787010536, 0.386062065605074, 0.564017448574305, 0.322985239792615, 0.975762875750661, 0.975396120222285, 0.169868181459606, 0.330435226438567, 0.222760551376268, 0.920518240425736, 0.035008690552786, 0.218932213494554, 0.592408939730376, 0.00278894533403218, 0.635503679513931, 0.628594990586862, 0.365686021279544, 0.554026905214414, 0.0434484113939106, 0.72073976858519, 0.701227918034419, 0.0538373365998268, 0.889852226478979, 0.167486419901252, 0.727622977690771, 0.0620911750011146, 0.56520706997253, 0.983019322389737, 0.342780474806204, 0.706210306147113, 0.751540750497952, 0.939613020978868, 0.714872799813747, 0.858665972016752, 0.466435466194525, 0.0530562228523195, 0.98864013585262, 0.489785948302597, 0.0587925303261727, 0.0655392522457987, 0.491564757423475, 0.710096965311095, 0.0512285411823541, 0.344587865518406, 0.828338026767597, 0.940060136374086, 0.235032749129459, 0.0915180779993534, 0.591598260914907, 0.231738155940548, 0.159363130573183, 0.453867607749999, 0.16970089240931, 0.346662797732279, 0.514075276907533, 0.113115215906873, 0.794155379291624, 0.465960283763707, 0.536217013373971, 0.632698511006311, 0.0971791008487344, 0.112759615061805, 0.997670748736709, 0.444674184545875, 0.690711392555386, 0.29733512038365, 0.327830488327891, 0.290128918131813, 0.699055283330381, 0.598597211763263, 0.617103482829407, 0.833892764290795, 0.862030499381945, 0.788969451095909, 0.200002840487286, 0.59690699679777, 0.852852193173021, 0.77488635876216, 0.396548870485276, 0.727565009146929, 0.0380043056793511, 0.0179611134808511, 0.939031084533781, 0.884045466082171, 0.88717925734818, 0.923344333656132, 0.398724063765258, 0.564525780268013, 0.450825272593647, 0.630715363891795, 0.679630458122119, 0.337349354289472, 0.609031104715541, 0.218997576041147, 0.172772050835192, 0.74137018318288, 0.26848580269143, 0.5296022715047, 0.963822784600779, 0.262877467786893, 0.669389872811735, 0.00863781245425344, 0.449463770957664, 0.156799361109734, 0.37732801027596, 0.307857068255544, 0.615539798745885, 0.156008323654532, 0.384217028040439, 0.889987265923992, 0.718475145986304, 0.00383056607097387, 0.689302477519959, 0.0217442931607366, 0.929198469733819, 0.936437619384378, 0.903860071673989, 0.65716495225206, 0.99030254338868, 0.134911378845572, 0.559262838447466, 0.721402778755873, 0.922358860261738, 0.901014847215265, 0.531697791768238, 0.315142219420522, 0.226151349721476, 0.229893177049235, 0.997016906738281, 0.570502262562513, 0.452379614580423, 0.169522803509608, 0.0260737927164882, 0.711924958974123, 0.512294691521674, 0.151980576105416, 0.590581369353458, 0.739606318529695, 0.925849438644946, 0.579393668565899, 0.695441169664264, 0.497626541880891, 0.99494175077416, 0.0521624418906868, 0.391923960763961, 0.804907829500735, 0.210215473081917, 0.620434299809858, 0.528598433360457, 0.589381282217801, 0.39621424744837, 0.243875481421128, 0.251553954556584, 0.521599346539006, 0.28670074720867, 0.0798587484750897, 0.509441755944863, 0.111488102003932, 0.854755953885615, 0.019756791414693, 0.722851952770725, 0.502647950546816, 0.818255105754361, 0.639862818643451, 0.0122086880728602, 0.600632347166538, 0.428527832729742, 0.437056360999122, 0.334908961551264, 0.949384881881997, 0.255237765843049, 0.405106090474874, 0.722924795001745, 0.621189028490335, 0.182109049288556, 0.368149726185948, 0.94568322179839, 0.866518256720155, 0.298479396151379, 0.732240300858393, 0.3859059184324, 0.935398494126275, 0.806167671224102, 0.269142012577504, 0.728615040658042, 0.331674934364855, 0.366458973614499, 0.139278387883678, 0.252871279604733, 0.352896349970251, 0.390659880125895, 0.598071999615058, 0.953899058047682, 0.613294249633327, 0.300163249019533, 0.435478755272925, 0.0420257111545652, 0.164659812580794, 0.565577914938331, 0.983678732067347, 0.513209352502599, 0.736047255340964, 0.739747437182814, 0.832139089470729, 0.561624512309209, 0.555755547014996, 0.557381593855098, 0.167744446545839, 0.809383478481323, 0.289349137572572, 0.264746084110811, 0.681942521361634, 0.728057092987001, 0.950338045600802, 0.547716724686325, 0.881459139520302, 0.737805681768805, 0.0339960234705359, 0.326373940100893, 0.957917264895514, 0.239273958606645, 0.0727325782645494, 0.21204410889186, 0.172582052415237, 0.0425429879687726, 0.309497874230146, 0.352242920780554, 0.668800784507766, 0.710520572727546, 0.0532854842022061, 0.127259612316266, 0.842302145203575, 0.575273009715602, 0.393095641862601, 0.751393000595272, 0.416298817843199, 0.338876759167761, 0.68499574996531, 0.000901202205568552, 0.943134748144075, 0.551692689303309, 0.910351835656911, 0.501629460835829, 0.732698698993772, 0.410277633462101, 0.787864493206143, 0.687597651034594, 0.195485973032191, 0.915639137616381, 0.64169338112697, 0.345722015248612, 0.475431907689199, 0.732194046489894, 0.760789141291752, 0.778628224972636, 0.248438182286918, 0.117258273996413, 0.0876029499340802, 0.947988289641216, 0.522214091615751, 0.875295263482258, 0.985481310635805, 0.643644207157195, 0.671190473949537, 0.611166608752683, 0.612657757475972, 0.770579006988555, 0.491519961273298, 0.74689839174971, 0.767093401867896, 0.661891464376822, 0.668282834580168, 0.462697779061273, 0.502783368108794, 0.819072797894478, 0.918618983356282, 0.567244807491079, 0.783860496710986, 0.955452984198928, 0.344559442484751, 0.956274406285957, 0.498898916179314, 0.834825120866299, 0.481286147609353, 0.0164828856941313, 0.74861382227391, 0.242579116253182, 0.35255073918961, 0.173258241266012, 0.556085417512804, 0.0717467188369483, 0.566019975580275, 0.70525703090243, 0.660511023597792, 0.353579012444243, 0.310998754575849, 0.28929511946626, 0.18272226373665, 0.959299902897328, 0.283279144205153, 0.710426769219339, 0.255325560225174, 0.625634801574051, 0.164750751806423, 0.727504286216572, 0.356800643261522, 0.384920162381604, 0.443132791668177, 0.104902333812788, 0.770003983285278, 0.504889358067885, 0.890068538254127, 0.426224133232608, 0.348285656189546, 0.866695565637201, 0.598322804551572, 0.655955640599132, 0.904631830286235, 0.309457858325914, 0.297382195247337, 0.194911315105855, 0.455676203593612, 0.853952818550169, 0.0846300104167312, 0.949998661410064, 0.75923627614975, 0.288449655519798, 0.518258317373693, 0.0440436352510005, 0.466034676646814, 0.259433424565941, 0.590274523710832, 0.42065303795971, 0.975477099185809, 0.809102650498971, 0.923607888864353, 0.0578176185954362, 0.169068890856579, 0.671027929754928, 0.0976256660651416, 0.0885143098421395, 0.0464478796347976, 0.373820488806814, 0.80507166380994, 0.834240542724729, 0.0283671417273581, 0.344952651532367, 0.385444259271026, 0.0658402021508664, 0.766829220810905, 0.398362694308162, 0.027682643616572, 0.951310241129249, 0.642041126731783, 0.114922658074647, 0.13947203475982, 0.653682932024822, 0.348135160515085, 0.136931270128116, 0.65758918877691, 0.185723824426532, 0.426218761363998, 0.84528597118333, 0.220868896692991, 0.0998379327356815, 0.81920458888635, 0.113279861863703, 0.130781644256786, 0.997019570320845, 0.802808337146416, 0.805404623970389, 0.88267492223531, 0.109581088181585, 0.367183362366632, 0.672581460559741, 0.0962923995684832, 0.862836296437308, 0.245159237878397, 0.776878285920247, 0.789389377227053, 0.136702844873071, 0.337275065714493, 0.2743968621362, 0.662804523250088, 0.632273013936356, 0.895028713159263, 0.408114088466391, 0.252289359923452, 0.956403937889263, 0.164391676196828, 0.249360797461122, 0.667660741135478), actual = c(0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.757677875459194, 0.886785438284278, 0.141607341356575, 0.42693106434308, 0.93577242968604, 0.190626195399091, 0.740459038875997, 0.173652331344783, 0.977095055626705, 0.0445604508277029, 0.442715292796493, 0.875594909535721, 0.771145114675164, 0.292386207263917, 0.835266506765038, 0.418820730177686, 0.662025360623375, 0.0636788383126259, 0.720419263700023, 0.66367976856418, 0.866646288428456, 0.648159973789006, 0.186875873245299, 0.119801702909172, 0.0552093137521297, 0.00321638397872448, 0.840131241828203, 0.843502046074718, 0.417564623290673, 0.875919286394492, 0.566779672866687, 0.871440407587215, 0.926733712665737, 0.798569977516308, 0.0307489442639053, 0.879368397872895, 0.429561521159485, 0.00648829550482333, 0.962964970385656, 0.443950238637626, 0.49403316504322, 0.0723632918670774, 0.482682007597759, 0.288589754607528, 0.68562391330488, 0.851709985407069, 0.39394729048945, 0.599904165137559, 0.858160015661269, 0.867019997676834, 0.780087533174083, 0.712064694613218, 0.515271069016308, 0.205724271945655, 0.988720086868852, 0.211618940345943, 0.83132230816409, 0.486877650022507, 0.981547251576558, 0.875635784352198, 0.91379888379015, 0.241271760547534, 0.0458681571763009, 0.269457328366116, 0.919946747133508, 0.582616593223065, 0.511752304853871, 0.449904330307618, 0.915223966352642, 0.719645577482879, 0.853339467663318, 0.923325644573197, 0.963063230272382, 0.355868442682549, 0.299326919252053, 0.393456655321643, 0.128276676638052, 0.671154468087479, 0.717012269422412, 0.00368289416655898, 0.747863959288225, 0.0139795870054513, 0.604541382286698, 0.267955714138225, 0.805394058348611, 0.31111350771971, 0.605873567285016, 0.782016509678215, 0.779570162063465, 0.240415553329512, 0.610149676445872, 0.391080020926893, 0.309620217652991, 0.803902533836663, 0.825432519195601, 0.341747138183564, 0.593768677441403, 0.31440788670443, 0.851873816922307, 0.815340073546395, 0.780520220985636, 0.64844142110087, 0.942723190644756, 0.289286843733862, 0.187861409736797, 0.372054317034781, 0.633834244450554, 0.578699056059122, 0.44782942882739, 0.358860695036128, 0.530828549060971, 0.441863328684121, 0.588642036309466, 0.607164284447208, 0.25883554155007, 0.92515658820048, 0.786021576263011, 0.786756256828085, 0.58807754307054, 0.859969342360273, 0.258410238195211, 0.154184328857809, 0.880544657353312, 0.833064875565469, 0.773566488176584, 0.109278619755059, 0.690888325683773, 0.549442224437371, 0.986542501021177, 0.988397876033559, 0.268253011861816, 0.724548411555588, 0.664621376199648, 0.641746422275901, 0.359220604412258, 0.363110254285857, 0.261961311334744, 0.690668890718371, 0.384501637192443, 0.0818185114767402, 0.0178982252255082, 0.37022127979435, 0.875190573045984, 0.571350436424837, 0.669930144678801, 0.333389577222988, 0.414941083174199, 0.552453263662755, 0.0237157004885375, 0.0189933255314827, 0.894209924852476, 0.44027700368315, 0.495589175960049, 0.499936250969768, 0.5927878706716, 0.424000134225935, 0.592207122361287, 0.266034686705098, 0.599287804914638, 0.0972075168974698, 0.169199499767274, 0.68911211611703, 0.867927952902392, 0.337914911331609, 0.943123231641948, 0.68532138178125, 0.929534775903448, 0.016274351393804, 0.275255408370867, 0.972054119454697, 0.621666952269152, 0.539354973938316, 0.512348112650216, 0.106200328795239, 0.372860830975696, 0.834394519217312, 0.0450633871369064, 0.903496867977083, 0.539007692597806, 0.784876184538007, 0.847231528488919, 0.825809124391526, 0.766114224214107, 0.491657049860805, 0.569818411255255, 0.361188417766243, 0.257174337282777, 0.435963063733652, 0.0150172738358378, 0.0399337350390851, 0.529199247946963, 0.115329078631476, 0.673322848742828, 0.682808166136965, 0.979600252117962, 0.628125763731077, 0.160147249232978, 0.916896119946614, 0.554255909286439, 0.823605513665825, 0.290799291105941, 0.913781145820394, 0.693356932373717, 0.268902870360762, 0.532055534422398, 0.422087135724723, 0.41770035889931, 0.476791457040235, 0.735029978211969, 0.21765469526872, 0.328219781629741, 0.950337389949709, 0.382570527959615, 0.258287174627185, 0.604335384676233, 0.781824119156227, 0.351327297976241, 0.927098049316555, 0.12027527904138, 0.128806501626968, 0.835095867747441, 0.363905702251941, 0.00742963841184974, 0.240101455477998, 0.826254279585555, 0.705814534565434, 0.769159604562446, 0.797765793744475, 0.688196272356436, 0.187019077362493, 0.406653525307775, 0.519003139575943, 0.757539463229477, 0.00434378162026405, 0.903961737873033, 0.688753167632967, 0.93747634999454, 0.117482106667012, 0.819626042619348, 0.177475093398243, 0.746308818925172, 0.186207920080051, 0.236289657885209, 0.0614344065543264, 0.229488577693701, 0.843513169325888, 0.937327898107469, 0.831509387120605, 0.0616343514993787, 0.821951491292566, 0.55695250350982, 0.15259167086333, 0.578071759780869, 0.740365249337628, 0.239881646120921, 0.0386123021598905, 0.988637284841388, 0.292723040562123, 0.0717724880669266, 0.0156218472402543, 0.67673830408603, 0.715216795681044, 0.155390459112823, 0.764791161520407, 0.531601804774255, 0.752496059518307, 0.589438590686768, 0.441868516616523, 0.420062012504786, 0.437444065930322, 0.661423444980755, 0.698461063904688, 0.449258283013478, 0.185183231951669, 0.191723951138556, 0.154499954776838, 0.664059634087607, 0.519191436003894, 0.594342046417296, 0.54467928269878, 0.913250953191891, 0.44436700688675, 0.257808748167008, 0.599423671141267, 0.0777589895296842, 0.0547848152928054, 0.633414096431807, 0.262191564776003, 0.596246625762433, 0.414494353812188, 0.327980850357562, 0.609812585636973, 0.737043516244739, 0.798119988525286, 0.942441428545862, 0.801845021313056, 0.97761986637488, 0.361639738548547, 0.0398120225872844, 0.0281557266134769, 0.843369909562171, 0.612541838781908, 0.484369560843334, 0.404889034340158, 0.704195348545909, 0.984391204547137, 0.044102776562795, 0.772666837321594, 0.682887594448403, 0.0478449396323413, 0.0546543586533517, 0.655314767034724, 0.600480042863637, 0.138205365277827, 0.233050630195066, 0.79723553545773, 0.218255328014493, 0.232312995241955, 0.202350527048111, 0.656621998175979, 0.872229611035436, 0.623447512509301, 0.823148943483829, 0.249086079886183, 0.476568008773029, 0.209810241591185, 0.925525150028989, 0.34456963185221, 0.278754604281858, 0.0337378419935703, 0.844594409689307, 0.293248044326901, 0.541218287777156, 0.606123176869005, 0.789322728756815, 0.231404192280024, 0.0221509586554021, 0.645927853416651, 0.442265219520777, 0.747004356468096, 0.0761984484270215, 0.707061217632145, 0.546181201236323, 0.256307627772912, 0.290883671026677, 0.128924444084987, 0.572931511094794, 0.633195599075407, 0.24719398166053, 0.632248993730173, 0.431247226661071, 0.125570291420445, 0.893709843978286, 0.665679384022951, 0.00703016994521022, 0.673511603381485, 0.893196809105575, 0.605254555586725, 0.259000951657072, 0.211698791244999, 0.995050325058401, 0.578267385484651, 0.616471158340573, 0.330338552594185, 0.630278300726786, 0.614864373346791, 0.636719174915925, 0.166612128261477, 0.308064420707524, 0.542100675869733, 0.597331484779716, 0.543656183872372, 0.838950449135154, 0.10152899986133, 0.68652528966777, 0.494046586332843, 0.343005969654769, 0.48273632931523, 0.683651696192101, 0.268942116294056, 0.520326284458861, 0.014219980686903, 0.834604872856289, 0.986220805440098, 0.777097248937935, 0.632922114105895, 0.083371106069535, 0.938459493452683, 0.866001638583839, 0.260313550010324, 0.74018865916878, 0.692335245199502, 0.665397029835731, 0.380318064009771, 0.248863375745714, 0.660549471154809, 0.491783089702949, 0.127947058295831, 0.0342360595241189, 0.530851654242724, 0.199633204843849, 0.931972217513248, 0.63305527693592, 0.458868904970586, 0.50673040561378, 0.38479535211809, 0.0867877763230354, 0.156308075645939, 0.0561556480824947, 0.70401846873574, 0.0622733868658543, 0.882933516288176, 0.0939170175697654, 0.772413757629693, 0.0036128347273916, 0.930488789454103, 0.215586809441447, 0.0586939034983516, 0.694449588190764, 0.896086030872539, 0.747922143433243, 0.0752038862556219, 0.81840525730513, 0.369260099017993, 0.599270771956071, 0.892198198242113, 0.227766977855936, 0.392343037528917, 0.00569845736026764, 0.349100998835638, 0.136139241047204, 0.99595110793598, 0.450751572847366, 0.572429618332535, 0.841214317129925, 0.506201746175066, 0.453454065136611, 0.57413431443274, 0.708189342636615, 0.417016871506348, 0.776671152561903, 0.196423122426495, 0.742217446910217, 0.80797086074017, 0.0402555267792195, 0.599700827850029, 0.0935325149912387, 0.335620556259528, 0.495502138277516, 0.401671243598685, 0.58941013738513, 0.640051222173497, 0.733628932619467, 0.926813311176375, 0.247883823467419, 0.778977774083614, 0.143202686682343, 0.836226801155135, 0.636852748924866, 0.598071582149714, 0.332117134705186, 0.591283764224499, 0.0615881234407425, 0.266683300724253, 0.899863025173545, 0.0703076082281768, 0.847465887432918, 0.864851800492033, 0.545262432890013, 0.0986966956406832, 0.222170650027692, 0.781512416666374, 0.945617287885398, 0.952697909669951, 0.220126411411911, 0.180800725705922, 0.0645502903498709, 0.155507545219734, 0.282386113656685, 0.579058362171054, 0.799920358462259, 0.469473487697542, 0.497787133557722, 0.068905828287825, 0.87206214456819, 0.373603886924684, 0.974718272453174, 0.424081000499427, 0.132723906077445, 0.773425163468346, 0.264597186585888, 0.584251104621217, 0.568673933157697, 0.449823025614023, 0.090133671881631, 0.113401398761198, 0.709874113323167, 0.915536449989304, 0.383965244749561, 0.467551176901907, 0.164059473434463, 0.60950938006863, 0.570763219147921, 0.98587155668065, 0.489911243319511, 0.52256254828535, 0.0666566996369511, 0.352740218164399, 0.14056362467818, 0.804850399959832, 0.125195137690753, 0.316027871565893, 0.27495794207789, 0.619566965615377, 0.000390607630833983, 0.198994558537379, 0.94459397578612, 0.417677576187998, 0.85410835291259, 0.24108477239497, 0.183324554469436, 0.0462999190203846, 0.786080090329051, 0.8831944570411, 0.0500851827673614, 0.177360107656568, 0.60603906866163, 0.831586273154244, 0.617659778567031, 0.681924977572635, 0.871473800390959, 0.0388676514849067, 0.386619827244431, 0.420669566141441, 0.392766098957509, 0.931641044327989, 0.214060131227598, 0.319876582827419, 0.698569959728047, 0.80011752131395, 0.455417758086696, 0.058889647712931, 0.593532589729875, 0.516663224203512, 0.983333291951567, 0.502033405005932, 0.878653676016256, 0.106012752745301, 0.110892746830359, 0.150894935010001, 0.933354685083032, 0.487150357104838, 0.932207207195461, 0.493360685650259, 0.880076081259176, 0.842748877825215, 0.0265516119543463, 0.140686692669988, 0.38221727963537, 0.371297359932214, 0.788919866317883, 0.897634642431512, 0.681997512700036, 0.364296429557726, 0.516508074942976, 0.0808898464310914, 0.995105822570622, 0.876998235238716, 0.714857609244063, 0.825963351642713, 0.0786768542602658, 0.989088577684015, 0.167958419537172, 0.322964695515111, 0.964275057427585, 0.929421340348199, 0.00450227363035083, 0.0960877281613648, 0.388509954791516, 0.794967976864427, 0.115976933855563, 0.228658419568092, 0.824194277403876, 0.587368183303624, 0.17714535468258, 0.685755344806239, 0.18981027463451, 0.206925306469202, 0.19737517205067, 0.086525633232668, 0.380896131973714, 0.705626812297851, 0.481297733960673, 0.296410120557994, 0.61780825140886, 0.687577073927969, 0.825511914910749, 0.953352293232456, 0.206929584499449, 0.33830670453608, 0.92660419899039, 0.190958003280684, 0.385384146589786, 0.573900222312659, 0.0909857356455177, 0.955934772733599, 0.932878624880686, 0.34059982188046, 0.416173846228048, 0.188474655384198, 0.477239467902109, 0.28534914413467, 0.950486169895157, 0.20657974248752, 0.239542165305465, 0.53877476323396, 0.204624491743743, 0.440740243997425, 0.19506452861242, 0.19898041151464, 0.30892585683614, 0.898866532137617, 0.920048335334286, 0.0707756518386304, 0.710338841192424, 0.0756711717694998, 0.454756783787161, 0.626177137019113, 0.849215385038406, 0.799352232832462, 0.901884766994044, 0.136770290322602, 0.795274837641045, 0.12775316904299, 0.425999708706513, 0.461904173484072, 0.909998151706532, 0.234096432337537, 0.374827768420801, 0.806621810654178, 0.751736780395731, 0.813230601837859, 0.657453276216984, 0.419069836148992, 0.497858886606991, 0.84265940845944, 0.337137145223096, 0.669502189615741, 0.832969600334764, 0.539437659550458, 0.74796479777433, 0.082968985196203, 0.704045055899769, 0.0485089470166713, 0.805159621639177, 0.734656075015664, 0.61417943239212, 0.539861346129328, 0.182647374458611, 0.502503138268366, 0.141453848220408, 0.331137602217495, 0.626640725182369, 0.563291357131675, 0.72124796663411, 0.115773125551641, 0.262880961643532, 0.81914984784089, 0.494519764790311, 0.83483445411548, 0.579715494532138, 0.903666536090896, 0.738453070865944, 0.964743540622294, 0.81457198294811, 0.327113034436479, 0.144293525489047, 0.0791573077440262, 0.934243723051623, 0.919528044527397, 0.721258474746719, 0.0284758927300572, 0.883725801715627, 0.109927661484107, 0.134205905953422, 0.0131019202526659, 0.254354294156656, 0.181044511962682, 0.628998259780928, 0.446530532324687, 0.253979607252404, 0.638698099413887, 0.869618676137179, 0.709115841891617, 0.568001409294084, 0.376080472720787, 0.663264772854745, 0.769895981065929, 0.466779781039804, 0.0378103726543486, 0.353013989515603, 0.624372727470472, 0.260829238919541, 0.694786655716598, 0.373421774478629, 0.341038774233311, 0.998609764035791, 0.500974465627223, 0.484202108345926, 0.916139962617308, 0.809578193817288, 0.0335967137943953, 0.784932906972244, 0.421033581253141, 0.91930571408011, 0.908103497466072, 0.398284140741453, 0.873855626443401, 0.564556993311271, 0.209996167803183, 0.906730074901134, 0.359870634740219, 0.136639723554254, 0.699951702263206, 0.427287899656221, 0.0323730257805437, 0.712209359044209, 0.328512101201341, 0.851200119126588, 0.573911869432777, 0.731551917968318, 0.681579133262858, 0.891970625147223, 0.457116437144578, 0.352647705934942, 0.558428784133866, 0.19471713854, 0.491745444014668, 0.113258018391207, 0.168705556774512, 0.78701478568837, 0.983511618105695, 0.278719539521262, 0.983378905104473, 0.069983446970582, 0.831039670156315, 0.030112313805148, 0.574614039389417, 0.869842906482518, 0.902957929531112, 0.885627949843183, 0.793900107266381, 0.681502216961235, 0.754105068510398, 0.76393755688332, 0.313823573756963, 0.685448449803516, 0.827368076192215, 0.406223949976265, 0.947357797063887, 0.76322262478061, 0.261531663825735, 0.0549832638353109, 0.0103248332161456, 0.519596894970164, 0.470752457855269, 0.26227067806758, 0.399215884273872, 0.488425018731505, 0.702989944955334, 0.731141187949106, 0.922982872230932, 0.398495957953855, 0.186078862054273, 0.788307565264404, 0.946039543254301, 0.81250591785647, 0.364584425929934, 0.528675930108875, 0.681151333963498, 0.773125667823479, 0.239806211553514, 0.469046369893476, 0.228874792112038, 0.549758973298594, 0.368821360403672, 0.430793790379539, 0.64991747494787, 0.2447096016258, 0.0161988055333495, 0.941110278014094, 0.155080265831202, 0.457076095975935, 0.0704496360849589, 0.0815200945362449, 0.797922224272043, 0.38015404692851, 0.163717618444934, 0.808678776724264, 0.353714502416551, 0.223827101988718, 0.0428670300170779, 0.456727962009609, 0.941759415669367, 0.00985462730750442, 0.8168242033571, 0.4096406288445, 0.910578161943704, 0.307839730987325, 0.957485322840512, 0.991972404764965, 0.579539542086422, 0.867385243531317, 0.824734848691151, 0.976815982954577, 0.381804974284023, 0.342841200064868, 0.937722494360059, 0.369214932434261, 0.275118780322373, 0.095239132642746, 0.745721828890964, 0.789342078613117, 0.458134799730033, 0.389227798674256, 0.561413864372298, 0.854413954075426, 0.596242662752047, 0.37328122346662, 0.911056025186554, 0.0912896066438407, 0.579180598957464, 0.617156635504216, 0.0657453646417707, 0.191751820035279, 0.258748002583161, 0.964465731754899, 0.544522703392431, 0.673756713978946, 0.00190587341785431, 0.634888595668599, 0.199266465613618, 0.482076623244211, 0.743374770274386, 0.219822022598237, 0.151548139052466, 0.733074776129797, 0.483497549314052, 0.900904637761414, 0.629454223439097, 0.377356455894187, 0.370549688348547, 0.878936722874641, 0.641283278353512, 0.972895517246798, 0.456508141476661, 0.40229638828896, 0.633456615498289, 0.270546418149024, 0.135058304527774, 0.656134965829551, 0.29292698064819, 0.641279429662973, 0.308026404585689, 0.191780095687136, 0.240836718119681, 0.0414808224886656, 0.306252938695252, 0.569358982145786, 0.377026580739766, 0.99410590226762, 0.371359638404101, 0.859716628678143, 0.969494373071939, 0.356500911759213, 0.220471377950162, 0.222719432553276, 0.731251051416621, 0.782035372685641, 0.0722346107941121, 0.0262685667257756, 0.12057616468519, 0.734148259041831, 0.450835446128622, 0.186449408298358, 0.043716095155105, 0.862567938165739, 0.398191889515147, 0.120396329788491, 0.298286176053807, 0.516672312514856, 0.776574716204777, 0.386832155287266, 0.52137358719483, 0.326064293505624, 0.546939166029915, 0.484211595728993, 0.0654341150075197, 0.767134893685579, 0.329863501945511, 0.62892308854498, 0.146855243248865, 0.419173487927765, 0.530443836469203, 0.894302858971059, 0.15028751292266, 0.0780820795334876, 0.93626306229271, 0.000362335937097669, 0.164352066814899, 0.684590467950329, 0.576025357469916, 0.103223343845457, 0.41996326809749, 0.934107200009748, 0.261507071554661, 0.259249942144379, 0.833424728596583, 0.405723571777344, 0.554467609152198, 0.15444835764356, 0.860018726438284, 0.272705169860274, 0.976730215130374, 0.868110889568925, 0.102221921784803, 0.546004518866539, 0.396270335186273, 0.539865121943876, 0.62862779176794, 0.962470661150292, 0.975648261141032, 0.58348107826896, 0.173143269727007, 0.837181226117536, 0.38947567413561, 0.0160795676056296, 0.361543036298826, 0.078419973840937, 0.611179223749787, 0.229358153650537, 0.140213570557535, 0.948086467105895, 0.868507030420005, 0.985043692868203, 0.283864174736664, 0.406109204050153, 0.143211856484413, 0.855823947815225, 0.031483092578128, 0.310628553386778, 0.248731769854203, 0.0860826657153666, 0.851408273447305, 0.740552892675623, 0.425872813444585, 0.275086212903261, 0.517866134643555, 0.479758818866685, 0.676632568938658, 0.191562835825607, 0.612416309304535, 0.276089600287378, 0.67835382767953, 0.766123159555718, 0.877677436219528, 0.0934905190952122, 0.721503847278655, 0.0354113022331148, 0.567099567735568, 0.648828000994399, 0.493623630376533, 0.417800132883713, 0.126182471867651, 0.0857811789028347, 0.771924273110926, 0.993754911934957, 0.395903274184093, 0.239020884269848, 0.70673173526302, 0.828870727447793, 0.610661410260946, 0.838972606929019, 0.071895775385201, 0.483270874246955, 0.108371314359829, 0.823358134832233, 0.233019387116656, 0.102092762710527, 0.820904727093875, 0.69209905480966, 0.969007768435404, 0.941505566937849, 0.059188703307882, 0.918399729998782, 0.223246122244745, 0.547295352444053, 0.0712268422357738, 0.665576644940302, 0.555899379309267), actual = c(0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"