Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.5922407        0.5910760 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5470598        0.5470598        0.5228683 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.248411325039342, 0.583947677630931, 0.0375236731488258, 0.755811451701447, 0.531974433455616, 0.387342452304438, 0.143795975251123, 0.773048016242683, 0.933102874318138, 0.714343461906537, 0.847047765972093, 0.355679211905226, 0.599541572621092, 0.117776277009398, 0.1477583181113, 0.852920213248581, 0.45939747360535, 0.646027540788054, 0.458367994055152, 0.681913270847872, 0.401540462160483, 0.207941699540243, 0.884390809107572, 0.0795138985849917, 0.438690474256873, 0.0190498817246407, 0.897747040726244, 0.823631607927382, 0.852542049251497, 0.804495927179232, 0.698709114920348, 0.509310537250713, 0.782645460451022, 0.0657677054405212, 0.351638957625255, 0.11506173433736, 0.987076253863052, 0.777233300032094, 0.694922071415931, 0.946995078818873, 0.157175140455365, 0.726256167516112, 0.82758108782582, 0.582925403257832, 0.0962312675546855, 0.890849341172725, 0.894930504960939, 0.833893153816462, 0.822943103034049, 0.193286051508039, 0.366644469555467, 0.699258527951315, 0.718289319192991, 0.376768455374986, 0.814826112007722, 0.513222840614617, 0.500095427734777, 0.767026238609105, 0.772930405568331, 0.714872814947739, 0.847467860439792, 0.268818094395101, 0.0649955742992461, 0.964469242608175, 0.85790813784115, 0.358147003920749, 0.577078417642042, 0.434114102041349, 0.1675966065377, 0.0758038456551731, 0.915161963552237, 0.034200674155727, 0.0058100048918277, 0.439674131572247, 0.2620682886336, 0.673903270624578, 0.504028142197058, 0.734891495201737, 0.673516172915697, 0.813010827638209, 0.50714935711585, 0.899165942566469, 0.835164607968181, 0.483357287012041, 0.575981076341122, 0.660813633352518, 0.0543300344143063, 0.53651696909219, 0.507204035529867, 0.984184002969414, 0.222325323149562, 0.378793355310336, 0.19127187365666, 0.418247333262116, 0.883033222518861, 0.0717683474067599, 0.929892185376957, 0.815341708017513, 0.0352145894430578, 0.824072666931897, 0.468532597646117, 0.00334439473226666, 0.321425150847062, 0.340049432823434, 0.347146160900593, 0.145883503835648, 0.360502209048718, 0.766555352834985, 0.772487242473289, 0.501223359722644, 0.953160065924749, 0.0228444905951619, 0.136787875555456, 0.314848536392674, 0.921866297489032, 0.0562490688171238, 0.522664275253192, 0.925303538097069, 0.000713887624442577, 0.523921852000058, 0.36295987223275, 0.543978051748127, 0.692175997653976, 0.212015442550182, 0.100887153530493, 0.0486010820604861, 0.253471219912171, 0.35010138200596, 0.389647591393441, 0.716197062283754, 0.232961971545592, 0.950206704204902, 0.270296352682635, 0.0998406941071153, 0.948399825487286, 0.861415536375716, 0.407926180167124, 0.0223744420800358, 0.717795525677502, 0.55326272896491, 0.961831542663276, 0.0278355407062918, 0.347717732423916, 0.656819421797991, 0.894474114524201, 0.673462721519172, 0.290457469876856, 0.997927806805819, 0.772912652464584, 0.57750340201892, 0.119993882253766, 0.0416753159370273, 0.874470607610419, 0.612652306444943, 0.80108444695361, 0.295108920661733, 0.205331191420555, 0.912526656640694, 0.83720924006775, 0.981044279411435, 0.38971110037528, 0.832447109278291, 0.100681807845831, 0.62425586162135, 0.911015849094838, 0.429396849358454, 0.63290617777966, 0.619673765031621, 0.365528946975246, 0.344094650819898, 0.605869329767302, 0.79800779139623, 0.474666586145759, 0.454767703311518, 0.451490819454193, 0.8352641670499, 0.974000244867057, 0.760197359137237, 0.118222658755258, 0.96384791447781, 0.0252923676744103, 0.207145230844617, 0.469610143918544, 0.849975531920791, 0.0492341117933393, 0.544740056851879, 0.909521793015301, 0.19056711695157, 0.048041061963886, 0.163717462914065, 0.98148065013811, 0.324109024368227, 0.188045673072338, 0.125636735465378, 0.84780315333046, 0.590266644721851, 0.522478009574115, 0.0337535366415977, 0.243948095245287, 0.137841128511354, 0.383344744797796, 0.543127661570907, 0.0276910024695098, 0.111048546852544, 0.398706194711849, 0.872396312886849, 0.137842393713072, 0.895102598238736, 0.92048073373735, 0.503569817868993, 0.752021580468863, 0.639200972625986, 0.00802553864195943, 0.45879730489105, 0.556576246861368, 0.211126553593203, 0.539864946389571, 0.793871389003471, 0.706280228681862, 0.0831779702566564, 0.962041275342926, 0.804345093201846, 0.0688067506998777, 0.775536336004734, 0.461339257191867, 0.842442248016596, 0.3456388306804, 0.829463595058769, 0.142026666551828, 0.574059836799279, 0.649220178136602, 0.678812744095922, 0.394829848781228, 0.951542860362679, 0.915880020242184, 0.475630320608616, 0.710580665851012, 0.675312436651438, 0.845635802717879, 0.904874868225306, 0.401131739839911, 0.220434414688498, 0.632959528360516, 0.595337535021827, 0.805361562175676, 0.815005180658773, 0.468366997782141, 0.639746053842828, 0.739586713025346, 0.0644286433234811, 0.669097460573539, 0.450949225341901, 0.258438089629635, 0.505631661508232, 0.598534188698977, 0.415277934633195, 0.514961258042604, 0.386635154951364, 0.618357654893771, 0.816863094922155, 0.951768678845838, 0.394015946891159, 0.108423128025606, 0.894717463990673, 0.163914396194741, 0.781338550616056, 0.380143345333636, 0.334844384808093, 0.58582198084332, 0.353094548219815, 0.448386487318203, 0.932366461725906, 0.908765987725928, 0.548136753495783, 0.0916918586008251, 0.466376860626042, 0.645552058238536, 0.944559482159093, 0.519383447244763, 0.519192770821974, 0.881396903190762, 0.761875847820193, 0.276213224511594, 0.982688682386652, 0.591138540999964, 0.952183080371469, 0.260659250197932, 0.395805928390473, 0.340623111929744, 0.513206238625571, 0.867411713814363, 0.0776441507041454, 0.230230242013931, 0.972247469937429, 0.299004317028448, 0.0770280011929572, 0.969448128249496, 0.0653107368852943, 0.723100802162662, 0.707617751555517, 0.182076629949734, 0.402976139448583, 0.872546856291592, 0.907523103523999, 0.318260821281001, 0.1549057723023, 0.00173064460977912, 0.081365013262257, 0.516769013600424, 0.939856463344768, 0.941391153726727, 0.112067443085834, 0.453061328735203, 0.824662290513515, 0.411084830993786, 0.214253267738968, 0.478993035620078, 0.874357547145337, 0.186535316286609, 0.902088404865935, 0.541405625408515, 0.512057266198099, 0.487090523354709, 0.171618348220363, 0.35472720162943, 0.579143770504743, 0.876936499495059, 0.37290498171933, 0.626992071047425, 0.60847221291624, 0.412118836771697, 0.283182118320838, 0.439928319305182, 0.596533410483971, 0.856778478948399, 0.746726663783193, 0.566342957317829, 0.714951337780803, 0.266152857569978, 0.85841691843234, 0.342260789591819, 0.57730356301181, 0.318473133258522, 0.359465541085228, 0.857550168177113, 0.580052003031597, 0.770738722989336, 0.867123477160931, 0.873652465175837, 0.0193513459526002, 0.967990765115246, 0.886772529222071, 0.22794521600008, 0.451048579066992, 0.656184669351205, 0.722685161978006, 0.780547116883099, 0.28637088695541, 0.702264361549169, 0.305725505342707, 0.613236933015287, 0.757110578939319, 0.999415307538584, 0.890856080921367, 0.119979130337015, 0.839427743339911, 0.64101890893653, 0.112438575830311, 0.0258255132939667, 0.065707677975297, 0.503232071874663, 0.123733989661559, 0.483751552412286, 0.123408404644579, 0.307616614503786, 0.442760009784251, 0.629333956399933, 0.451992500806227, 0.312275267904624, 0.0704746176488698, 0.0805496103130281, 0.298925471724942, 0.400908807525411, 0.194471212569624, 0.689710262231529, 0.597776820417494, 0.28456759522669, 0.15336929471232, 0.365151544334367, 0.49699036218226, 0.853853919077665, 0.729699397459626, 0.775808460311964, 0.775700341444463, 0.466951378621161, 0.508172735571861, 0.426615976728499, 0.352530322270468, 0.372475646203384, 0.377378951292485, 0.112409272929654, 0.544106566347182, 0.90637899399735, 0.924976133275777, 0.207099695457146, 0.29485954134725, 0.658441059989855, 0.580774256493896, 0.0683338600210845, 0.0650114770978689, 0.715821211691946, 0.847550232196227, 0.934000108158216, 0.236532491166145, 0.734325338620692, 0.696566494181752, 0.859362104441971, 0.979273914126679, 0.135187638457865, 0.184160399483517, 0.0743597205728292, 0.269050711998716, 0.0142002983484417, 0.0134600475430489, 0.159872216172516, 0.863073677290231, 0.850892248330638, 0.100164995063096, 0.346578327706084, 0.30906303296797, 0.77153358166106, 0.616566287120804, 0.755393302999437, 0.93206562823616, 0.410853398963809, 0.36842125793919, 0.387171134119853, 0.978033925173804, 0.122010152321309, 0.39092904701829, 0.343347070273012, 0.107234509894624, 0.122574881883338, 0.0433208520989865, 0.162023841403425, 0.281584445619956, 0.43854540400207, 0.284154756926, 0.370476001407951, 0.471717815846205, 0.878423249581829, 0.663761471398175, 0.554923807503656, 0.18048899830319, 0.970192708540708, 0.452274373499677, 0.282358056399971, 0.251620775554329, 0.385827977675945, 0.734545125626028, 0.410481734899804, 0.404838321264833, 0.60107638174668, 0.790885410271585, 0.794909994583577, 0.102856750832871, 0.44696091930382, 0.760955692268908, 0.528873759554699, 0.732443731278181, 0.926946414401755, 0.374938207911327, 0.514484743354842, 0.88565567927435, 0.862785523990169, 0.0167712648399174, 0.0350627722218633, 0.217121570603922, 0.506737145129591, 0.893174592871219, 0.0287025880534202, 0.516255399677902, 0.559890448348597, 0.726028651697561, 0.610134393908083, 0.591406462481245, 0.662980157881975, 0.799018931342289, 0.98416393995285, 0.0532668486703187, 0.687279752455652, 0.969390696613118, 0.252146865008399, 0.166201053652912, 0.313745815539733, 0.767315182834864, 0.31491813948378, 0.460957630071789, 0.674701002892107, 0.859113903483376, 0.431692323181778, 0.450325894868001, 0.441587419714779, 0.741946660215035, 0.199896670179442, 0.0393529438879341, 0.886852223891765, 0.569955810438842, 0.931176357204095, 0.00229759933426976, 0.945382088422775, 0.616223132004961, 0.356200214941055, 0.13845706731081, 0.956360066542402, 0.0441666904371232, 0.835330929374322, 0.676698875846341, 0.261100262403488, 0.561968204798177, 0.0895639173686504, 0.0703726143110543, 0.68367500975728, 0.326873200014234, 0.944738794350997, 0.215789546258748, 0.643390591489151, 0.106988428626209, 0.233599329832941, 0.928548583760858, 0.971709716599435, 0.804895299952477, 0.625272243982181, 0.60327370557934, 0.590078870300204, 0.210936840623617, 0.408639986300841, 0.610346358967945, 0.705019640736282, 0.287510497728363, 0.462709283456206, 0.159467013552785, 0.423567987279966, 0.744620428420603, 0.179432019824162, 0.0478395384270698, 0.26851744344458, 0.73804431501776, 0.430176519555971, 0.189861950697377, 0.776184991002083, 0.976454985328019, 0.229866749141365, 0.560865733539686, 0.475384168559685, 0.00565495737828314, 0.223061092430726, 0.468422091566026, 0.534700966440141, 0.968987839994952, 0.928106877719983, 0.768472793977708, 0.198205455206335, 0.33829842065461, 0.613943193806335, 0.517614383948967, 0.421310090692714, 0.259093896485865, 0.0345298319589347, 0.778397000627592, 0.154348124517128, 0.230612992076203, 0.368558651767671, 0.652285170275718, 0.439069502986968, 0.392033529700711, 0.185800700448453, 0.692426404217258, 0.133946313289925, 0.828157246578485, 0.693705741083249, 0.0467166325543076, 0.380895246053115, 0.181472100084648, 0.594533683964983, 0.696389674209058, 0.680686736712232, 0.343743088189512, 0.718507790705189, 0.286029927432537, 0.501415948849171, 0.449494425440207, 0.607327590929344, 0.670639100950211, 0.313706727698445, 0.890437691938132, 0.893248225795105, 0.443947559921071, 0.296466907020658, 0.354622996179387, 0.191270685056224, 0.556385564617813, 0.695720150833949, 0.785644834861159, 0.690699596656486, 0.268935499014333, 0.277129164431244, 0.760752506554127, 0.408573056338355, 0.265092794317752, 0.295550398528576, 0.522365542827174, 0.590472543844953, 0.591915148776025, 0.867877465905622, 0.725207226118073, 0.756259496090934, 0.982325227931142, 0.172079874901101, 0.393435470992699, 0.170073157642037, 0.687683725263923, 0.123650145484135, 0.611253440147266, 0.574786518700421, 0.0135363203007728, 0.276140704518184, 0.0427396211307496, 0.961491893278435, 0.362578486790881, 0.822495149681345, 0.543930110288784, 0.381258074892685, 0.293549906695262, 0.826350732706487, 0.483984013320878, 0.158196549629793, 0.862051696050912, 0.964529800461605, 0.0807283546309918, 0.181705672293901, 0.606596063822508, 0.531473074806854, 0.277183308266103, 0.869814603589475, 0.958684598561376, 0.416771102230996, 0.8859270690009, 0.916005530860275, 0.839686666615307, 0.548012824496254, 0.519476590212435, 0.332720919046551, 0.468795879278332, 0.994594468502328, 0.730381942586973, 0.495616072323173, 0.417184218997136, 0.0708951717242599, 0.365296540781856, 0.424071535235271, 0.701921579195186, 0.631786424200982, 0.568126516183838, 0.243614589096978, 0.952226146589965, 0.391549784457311, 0.491838308284059, 0.294634397840127, 0.710610299371183, 0.392015546327457, 0.508184621110559, 0.759977473411709, 0.422841068124399, 0.424717557383701, 0.68721809098497, 0.0259283813647926, 0.170349643100053, 0.397522376850247, 0.452578775584698, 0.978824390796944, 0.937768721254542, 0.950725406641141, 0.0957066924311221, 0.873129109852016, 0.708104506833479, 0.420653040986508, 0.509180726949126, 0.885097659192979, 0.289446498034522, 0.0196393511723727, 0.66838353080675, 0.83840254181996, 0.560908251674846, 0.53768095606938, 0.732823552330956, 0.625946729909629, 0.262758907629177, 0.170605232473463, 0.26310914568603, 0.159645651001483, 0.0296308191027492, 0.807972843293101, 0.0681991749443114, 0.913212446495891, 0.0772702782414854, 0.869660951895639, 0.0359906868543476, 0.300376656698063, 0.122910714941099, 0.00872769090346992, 0.744911554036662, 0.245383135974407, 0.309272323735058, 0.910740626743063, 0.302350386511534, 0.196828356711194, 0.256196737987921, 0.443158941110596, 0.662057167151943, 0.00356552097946405, 0.908303733449429, 0.347328826319426, 0.774748542811722, 0.129949742928147, 0.490389981307089, 0.444060048786923, 0.590106663526967, 0.016954030143097, 0.566774292849004, 0.226213439600542, 0.734885679092258, 0.694699565880001, 0.0639514720533043, 0.724554351065308, 0.0735563752241433, 0.466185885248706, 0.597605464747176, 0.0340769102331251, 0.929789450019598, 0.960517002502456, 0.773926412453875, 0.015882465057075, 0.340942949056625, 0.790654937503859, 0.522603287594393, 0.120483161415905, 0.248278132174164, 0.772826154250652, 0.925455153686926, 0.426022832514718, 0.987060151062906, 0.866986178793013, 0.434648559428751, 0.95735794515349, 0.285440018167719, 0.51974521856755, 0.559430969646201, 0.57159653538838, 0.649099062429741, 0.219977071275935, 0.13276217225939, 0.669070141157135, 0.90228283428587, 0.467741461237893, 0.34890342829749, 0.423927847761661, 0.0247216953430325, 0.463317264104262, 0.386001863051206, 0.36143462616019, 0.592896668240428, 0.478484214982018, 0.968301387969404, 0.122050635749474, 0.0125590688548982, 0.682802675524727, 0.196171860676259, 0.666093937586993, 0.990292116533965, 0.695012452779338, 0.874572356930003, 0.847076943842694, 0.424040860263631, 0.544261821079999, 0.492626557592303, 0.954332252265885, 0.235077981371433, 0.37330755032599, 0.249543052399531, 0.084724155953154, 0.536830275319517, 0.369872269919142, 0.0353422435000539, 0.968594621634111, 0.199532811995596, 0.994931001914665, 0.753136872081086, 0.897770722163841, 0.480343349976465, 0.602898978628218, 0.340971850790083, 0.225270795635879, 0.487464258214459, 0.171248581493273, 0.163643656065688, 0.771217385074124, 0.51822255179286, 0.0306551402900368, 0.120345345465466, 0.336779177421704, 0.892663420876488, 0.982156134210527, 0.856374245136976, 0.466161025688052, 0.630431813886389, 0.221244370797649, 0.213137964718044, 0.905094425426796, 0.631195431575179, 0.12284476030618, 0.957739078905433, 0.307808421552181, 0.843170819571242, 0.471887509338558, 0.890899134567007, 0.521933196345344, 0.941849012160674, 0.26362935337238, 0.419356361264363, 0.306141688255593, 0.646761980373412, 0.547541741281748, 0.226515052374452, 0.684540347661823, 0.377117840340361, 0.364549035206437, 0.690280218608677, 0.46573551138863, 0.336277605965734, 0.67721473891288, 0.838154865661636, 0.873897733399644, 0.544941590866074, 0.776099379872903, 0.291969997808337, 0.578267605276778, 0.235027260612696, 0.709365723421797, 0.856094760354608, 0.994611322646961, 0.872777509503067, 0.768424321431667, 0.855739073129371, 0.393308620899916, 0.60251923580654, 0.9032414003741, 0.462410690961406, 0.638096184004098, 0.125461676856503, 0.846225308720022, 0.534876801306382, 0.738227211171761, 0.924303823150694, 0.688080908730626, 0.622534451307729, 0.191763599636033, 0.609618011629209, 0.0397451238241047, 0.535822854144499, 0.745181326987222, 0.534003145294264, 0.349963532062247, 0.461736532161012, 0.871835491154343, 0.465826943283901, 0.188421563478187, 0.825801882194355, 0.198229635134339, 0.902795248664916, 0.22584642842412, 0.4133572189603, 0.106942285550758, 0.772431976860389, 0.295167764415964, 0.319547787308693, 0.99058056040667, 0.934363795910031, 0.771459465846419, 0.664338234113529, 0.318417818518355, 0.63230138970539, 0.798655604710802, 0.393765112850815, 0.225616774754599, 0.471004780149087, 0.0746387580875307, 0.219758916646242, 0.605659671593457, 0.944805737351999, 0.250778700923547, 0.0363172711804509, 0.674762901384383, 0.971657998859882, 0.523873530095443, 0.691356653114781, 0.28572972305119, 0.609437110600993, 0.752014661673456, 0.656632856465876, 0.906096989754587, 0.475904084043577, 0.676380413351581, 0.250147465849295, 0.0536444685421884, 0.275132891954854, 0.1270006892737, 0.253793898038566, 0.327949413796887, 0.121881347382441, 0.485357471974567, 0.16789314779453, 0.344539340119809, 0.139662228990346, 0.598028217907995, 0.54262484004721, 0.43432853743434, 0.294514378067106, 0.752564940834418, 0.666774678509682, 0.784486666787416, 0.365630135405809, 0.888595202937722, 0.431498263962567, 0.304985150927678, 0.842391088837758, 0.795181138208136, 0.550123737892136, 0.972641881089658, 0.949260296300054, 0.0392656978219748, 0.932009176583961, 0.724150577792898, 0.717624275712296, 0.05146181024611, 0.1646775866393, 0.224412615178153, 0.164217789657414, 0.690619031898677, 0.2987750985194, 0.126871851505712, 0.065146220382303, 0.673934670398012, 0.627677782671526, 0.931317725451663, 0.479731507599354, 0.441046723630279, 0.557209432125092, 0.608970885630697, 0.200516612036154, 0.602382455021143, 0.764177669771016, 0.54991103708744, 0.871672838227823, 0.945745197590441, 0.401427330216393, 0.894172907108441, 0.654367382172495, 0.180221292655915, 0.152051886310801, 0.601153831230476, 0.5956853677053, 0.950386423617601, 0.533411829266697, 0.0688144033774734, 0.360102472594008, 0.325560164405033, 0.0774529359769076, 0.130961118033156, 0.699831180740148, 0.425034112064168, 0.441447661491111, 0.187226236332208, 0.772418009117246, 0.411295177182183, 0.947012173011899, 0.182976626092568, 0.135529480176046, 0.725167819764465, 0.079278013901785, 0.96964915166609, 0.497926009586081, 0.455289585283026, 0.254060700070113, 0.291255419375375, 0.227913138922304, 0.940902689471841, 0.88988516991958, 0.0618087772745639, 0.365426213480532), actual = c(0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.631360710365698, 0.966625173343346, 0.215809455607086, 0.443098542746156, 0.990938915638253, 0.771942277206108, 0.0104690347798169, 0.691127787809819, 0.172841559397057, 0.745862603653222, 0.425281053641811, 0.37723698373884, 0.853369909105822, 0.823719373438507, 0.83114936738275, 0.950726371025667, 0.0533831359352916, 0.569197226548567, 0.785011266358197, 0.0206805630587041, 0.640693189110607, 0.234174238052219, 0.0930113871581852, 0.303503485163674, 0.200832405127585, 0.708828435046598, 0.809863854665309, 0.519521571230143, 0.278663577511907, 0.595035122940317, 0.876260602613911, 0.363035638816655, 0.560876451898366, 0.299889110028744, 0.504157940624282, 0.848826863337308, 0.785382004687563, 0.0769991860724986, 0.0593446127604693, 0.515728543978184, 0.743842388037592, 0.376399862812832, 0.0357321174815297, 0.0606775430496782, 0.276320581557229, 0.946533867623657, 0.00727525469847023, 0.16584536479786, 0.58947042748332, 0.482591483509168, 0.645937916822731, 0.942942042835057, 0.412580319447443, 0.669250637758523, 0.503851198824123, 0.10865898267366, 0.915769858751446, 0.169916647719219, 0.8388677588664, 0.701394123258069, 0.525792118161917, 0.533907340606675, 0.113945762859657, 0.252303563524038, 0.846230628667399, 0.649858865421265, 0.982951231766492, 0.402363038621843, 0.982331532984972, 0.459846304962412, 0.448075952706859, 0.336807889631018, 0.575373806990683, 0.205710785929114, 0.803217319305986, 0.722477700794116, 0.870045856107026, 0.660276331938803, 0.810746662784368, 0.929716998944059, 0.818993153050542, 0.743119929684326, 0.912162092514336, 0.191759053152055, 0.15202388074249, 0.631258768728003, 0.936542272567749, 0.550425426103175, 0.314552154159173, 0.935969396261498, 0.244328659027815, 0.775661538355052, 0.853655590442941, 0.139104459900409, 0.0133301881141961, 0.489107368979603, 0.142914078896865, 0.534765561344102, 0.670744086615741, 0.703605009242892, 0.244551426498219, 0.19708417262882, 0.711741115897894, 0.0922016720287502, 0.408957353793085, 0.0192403248511255, 0.323229684261605, 0.190336281666532, 0.296221148455516, 0.233707058941945, 0.697570481803268, 0.351427926914766, 0.611429566051811, 0.25381125533022, 0.245811516419053, 0.592264365404844, 0.758597081759945, 0.698764834320173, 0.0492348875850439, 0.897953026695177, 0.356427224352956, 0.947380736703053, 0.975952621316537, 0.915670193964615, 0.527730139205232, 0.758672166382894, 0.961857954273, 0.816810683114454, 0.0537903727963567, 0.0409512009937316, 0.504806542769074, 0.117667434038594, 0.465154864592478, 0.0126460890751332, 0.866381191881374, 0.966621315805241, 0.76204585051164, 0.275691373739392, 0.0814172052778304, 0.426359705859795, 0.240260284161195, 0.640187939861789, 0.997133844299242, 0.667435710085556, 0.840986229479313, 0.9347583882045, 0.0203344526235014, 0.0316929356195033, 0.877173167653382, 0.545365802245215, 0.848946355748922, 0.938637933228165, 0.218704375438392, 0.464568390510976, 0.022096945438534, 0.387146405642852, 0.390414563706145, 0.390195402316749, 0.849487397586927, 0.124821837292984, 0.978751102928072, 0.00449315551668406, 0.128802741644904, 0.549214680679142, 0.0524666986893862, 0.310797234065831, 0.151375731918961, 0.56651024823077, 0.243981410050765, 0.170878804754466, 0.743588174460456, 0.317687089089304, 0.416413606144488, 0.942590605700389, 0.401456052903086, 0.329057819442824, 0.0581476003862917, 0.0752620853018016, 0.862872173078358, 0.158288941252977, 0.0285238933283836, 0.73464593780227, 0.80691697797738, 0.320071771508083, 0.251003863289952, 0.954400616232306, 0.804029312217608, 0.0840529839042574, 0.54873084067367, 0.306539388839155, 0.115506738191471, 0.213440220803022, 0.515738462563604, 0.159878197358921, 0.706818320788443, 0.993485262617469, 0.973092631902546, 0.755540370475501, 0.765416581416503, 0.677021034993231, 0.512651183409616, 0.715712409466505, 0.971061324002221, 0.941934903617948, 0.805243174312636, 0.412404523929581, 0.324244824936613, 0.485521493945271, 0.462378249270841, 0.98377565969713, 0.433897255454212, 0.388988570077345, 0.70690478105098, 0.0861803088337183, 0.774784318404272, 0.945153023814783, 0.635896999854594, 0.255534408148378, 0.354879577411339, 0.34476128872484, 0.743925506481901, 0.297043661121279, 0.255747182061896, 0.240434453589842, 0.426588267087936, 0.286826686933637, 0.695020065642893, 0.0250565537717193, 0.303834382910281, 0.282528852112591, 0.0855340699199587, 0.459817900555208, 0.128000115975738, 0.993704301072285, 0.253960208967328, 0.320808540796861, 0.283142239553854, 0.119465221650898, 0.760200559627265, 0.570444902405143, 0.531392997363582, 0.695021510589868, 0.0537003397475928, 0.627667157910764, 0.745234698057175, 0.156168616842479, 0.504155553411692, 0.610603358829394, 0.335913925431669, 0.337433494394645, 0.078040705062449, 0.379361106315628, 0.213742478750646, 0.354541455861181, 0.587582048261538, 0.294884032569826, 0.495585585711524, 0.708156413864344, 0.562016301555559, 0.501906728139147, 0.424751533195376, 0.895894513465464, 0.714725367259234, 0.901610029162839, 0.925287948455662, 0.429376062937081, 0.298371249809861, 0.0154289123602211, 0.877167828846723, 0.785645978525281, 0.250778187531978, 0.287051623919979, 0.319462151033804, 0.700552279129624, 0.125589693197981, 0.970456037204713, 0.909891262184829, 0.395171738229692, 0.239930511219427, 0.301353324437514, 0.0129095329903066, 0.763607401400805, 0.562753011472523, 0.241466542938724, 0.0531503839883953, 0.945062131388113, 0.644826146773994, 0.818379604257643, 0.894864414352924, 0.372340278699994, 0.694236801238731, 0.53430881164968, 0.419657468097284, 0.498642332619056, 0.938901925925165, 0.216579359490424, 0.565254763467237, 0.785024291137233, 0.983009910909459, 0.0400988927576691, 0.986225760541856, 0.323769876267761, 0.0316826780326664, 0.0742985201068223, 0.261012125294656, 0.940626350464299, 0.968323861015961, 0.0303828387986869, 0.700997374253348, 0.310230270028114, 0.322561897104606, 0.795780493412167, 0.406110362382606, 0.934712584363297, 0.37643106142059, 0.0406543659046292, 0.719706353032961, 0.757764049107209, 0.906719512538984, 0.652895807987079, 0.919189307605848, 0.235271491808817, 0.496780381072313, 0.549212665762752, 0.858287279028445, 0.867816192097962, 0.377522374503314, 0.17079060152173, 0.181550872977823, 0.189726697281003, 0.342871768632904, 0.370441308477893, 0.350596302421764, 0.490968545665964, 0.48515733750537, 0.288365782937035, 0.274326692335308, 0.35817835945636, 0.733895358629525, 0.307070553069934, 0.684948228066787, 0.186080931453034, 0.84172765398398, 0.310701694339514, 0.656957120634615, 0.405186095274985, 0.300770737463608, 0.191199040273204, 0.225073762238026, 0.114711589645594, 0.658621376380324, 0.513686024583876, 0.264232547488064, 0.834019876318052, 0.712567124748603, 0.467052984517068, 0.889316980028525, 0.722701800754294, 0.828315198654309, 0.871585615212098, 0.0736895622685552, 0.836942958645523, 0.20158657990396, 0.85682030278258, 0.148575196275488, 0.31881558150053, 0.282450549304485, 0.245668119518086, 0.220736838178709, 0.640658778138459, 0.494856838835403, 0.981902271974832, 0.191369959386066, 0.436106197535992, 0.164458285551518, 0.504987387219444, 0.869232319993898, 0.616442485712469, 0.879034721525386, 0.414381675189361, 0.975630456814542, 0.295393423875794, 0.0782322273589671, 0.269169957144186, 0.968947675777599, 0.801489957375452, 0.957162785343826, 0.533528941916302, 0.342854139627889, 0.36868767789565, 0.860565523384139, 0.420674822758883, 0.762382792541757, 0.88602078310214, 0.282437167596072, 0.304061600239947, 0.408507217187434, 0.0435104165226221, 0.6214480525814, 0.596284204395488, 0.361102470895275, 0.251417031046003, 0.12514087697491, 0.07572365202941, 0.439547008601949, 0.0476430545095354, 0.741070669377223, 0.209607494762167, 0.771895283134654, 0.791451006662101, 0.712081496138126, 0.123532087542117, 0.410517835989594, 0.888587716734037, 0.345068725058809, 0.180495331995189, 0.789454356068745, 0.8198436726816, 0.643895864486694, 0.287662913091481, 0.397558529628441, 0.607200091006234, 0.765068110311404, 0.541620388627052, 0.855091928504407, 0.862658928381279, 0.272911610780284, 0.581857503624633, 0.871659967349842, 0.606563539477065, 0.431133883306757, 0.340288743842393, 0.613020039629191, 0.601309627993032, 0.765988881001249, 0.445963731501251, 0.698812118265778, 0.468074808362871, 0.53398988652043, 0.513191688107327, 0.126504431478679, 0.0811594012193382, 0.579141345340759, 0.0591884220484644, 0.383584201568738, 0.428011293290183, 0.248120805481449, 0.541640941053629, 0.532229486154392, 0.248649902874604, 0.105586079647765, 0.472430655034259, 0.372050470672548, 0.0137394743505865, 0.659194000763819, 0.816447202349082, 0.829883402446285, 0.040595552418381, 0.858902754960582, 0.706595188705251, 0.265910404967144, 0.194616799941286, 0.71023135422729, 0.266523799160495, 0.259076979709789, 0.247805092250928, 0.228109522722661, 0.469106980832294, 0.386532567441463, 0.996925962390378, 0.390194832114503, 0.522781672421843, 0.34881760366261, 0.684769411105663, 0.13539829174988, 0.333840194856748, 0.790767984697595, 0.418637174880132, 0.206466612638906, 0.611108207609504, 0.0880158164072782, 0.206956820795313, 0.165419747820124, 0.991811205865815, 0.372064825380221, 0.308683363022283, 0.563292127102613, 0.639505453407764, 0.316580067388713, 0.0269161658361554, 0.243578085908666, 0.272676299326122, 0.41617936338298, 0.544930169126019, 0.914797875331715, 0.794162690872326, 0.947082370053977, 0.305635706754401, 0.630643738666549, 0.800759019330144, 0.139999514911324, 0.180003981105983, 0.643721225904301, 0.437415386084467, 0.685235736658797, 0.267322205938399, 0.98986683995463, 0.547113103792071, 0.120291386265308, 0.429532387992367, 0.741995678516105, 0.025878332555294, 0.128478148952127, 0.784182994859293, 0.415011368924752, 0.344875396462157, 0.897872525965795, 0.655128711601719, 0.289435993647203, 0.77808873844333, 0.830189801054075, 0.621886006323621, 0.451614889316261, 0.32649463112466, 0.398486783262342, 0.957066952018067, 0.594200255116448, 0.736381245777011, 0.129868712509051, 0.850203013513237, 0.460174823412672, 0.450454897945747, 0.728374248370528, 0.706114165019244, 0.950815162388608, 0.988914174959064, 0.668285181280226, 0.669732343638316, 0.369483069516718, 0.903202944900841, 0.5288636891637, 0.90085136028938, 0.557017770130187, 0.287569547537714, 0.438359427964315, 0.902596888830885, 0.56966804410331, 0.421048010932282, 0.323692991631106, 0.120447830995545, 0.565351701108739, 0.805814953288063, 0.499137573176995, 0.993873656727374, 0.974841129034758, 0.226810871390626, 0.124285418074578, 0.787693685619161, 0.760441290680319, 0.164280337514356, 0.38471963442862, 0.728550295112655, 0.928035893477499, 0.924168966477737, 0.851278292248026, 0.013994136126712, 0.876183995744213, 0.336141481762752, 0.406477581942454, 0.0825079230125993, 0.914814956719056, 0.866841260110959, 0.334627866512164, 0.0791547049302608, 0.754319943953305, 0.635549053316936, 0.488510238472372, 0.416015064576641, 0.448851021239534, 0.394501993665472, 0.963494991185144, 0.410332784056664, 0.404147725086659, 0.598522916203365, 0.0628612190485001, 0.170920849312097, 0.490261303493753, 0.883397523779422, 0.428973878268152, 0.398305372102186, 0.654555126558989, 0.452023580204695, 0.752183297416195, 0.127062373561785, 0.723657076712698, 0.0370178325101733, 0.778799715684727, 0.679649741388857, 0.285274373134598, 0.988201139261946, 0.823137341300026, 0.370382437482476, 0.115479427622631, 0.790629312396049, 0.771902103209868, 0.543472850928083, 0.741099068662152, 0.0893715503625572, 0.673820433672518, 0.618068454321474, 0.884850719710812, 0.674889194080606, 0.317836537025869, 0.846503555541858, 0.99101046170108, 0.375117398798466, 0.651504972018301, 0.0155538758262992, 0.162939658155665, 0.913854593411088, 0.228767579887062, 0.421077454229817, 0.780007548630238, 0.717591175576672, 0.456054722657427, 0.170322840567678, 0.623201472684741, 0.785679748980328, 0.29331276519224, 0.626641892595217, 0.990915157133713, 0.817534355679527, 0.906486243242398, 0.0372304252814502, 0.532312657684088, 0.841014756588265, 0.25565215642564, 0.372913790633902, 0.768737743142992, 0.198796457378194, 0.430268527474254, 0.29539531073533, 0.705504314042628, 0.819415374659002, 0.917700334452093, 0.256835934240371, 0.181846479652449, 0.711146004032344, 0.890921431360766, 0.819233084563166, 0.821930073900148, 0.683524300577119, 0.138029570225626, 0.0195421737153083, 0.970249643316492, 0.235155731905252, 0.498243933077902, 0.610139201395214, 0.664686172734946, 0.115902761463076, 0.513530859490857, 0.120704457629472, 0.421065715374425, 0.335648633539677, 0.846278701676056, 0.887178875040263, 0.733056385302916, 0.0777490644250065, 0.279705002903938, 0.455035197781399, 0.257388429483399, 0.181349015561864, 0.0129558974877, 0.0893104576971382, 0.608250042423606, 0.0282917283475399, 0.760945251444355, 0.178625001572073, 0.127673140959814, 0.0867416427936405, 0.374626930803061, 0.120785369072109, 0.206009424291551, 0.220564960036427, 0.406701027182862, 0.0864328744355589, 0.608744679950178, 0.859928002581, 0.661948724417016, 0.456462856382132, 0.817301719915122, 0.0607173766475171, 0.929159773280844, 0.057881387649104, 0.44954706565477, 0.78571923985146, 0.425531281623989, 0.843988150591031, 0.887358298990875, 0.535758401500061, 0.478025087155402, 0.940920231863856, 0.647612965898588, 0.997293202439323, 0.993048907490447, 0.410455630160868, 0.565394966164604, 0.793040446238592, 0.486604421166703, 0.209046056494117, 0.773258443456143, 0.7194849466905, 0.564443127484992, 0.739069504896179, 0.708521461114287, 0.331882064696401, 0.565298047382385, 0.138219235930592, 0.356168524827808, 0.920167215401307, 0.609331177314743, 0.771103564882651, 0.752177964430302, 0.276060869451612, 0.133123196428642, 0.275130823254585, 0.430421571712941, 0.293884907150641, 0.874481842154637, 0.340491220122203, 0.790686600841582, 0.303467281861231, 0.864602952729911, 0.214327850844711, 0.46274042269215, 0.301775862229988, 0.0191055005416274, 0.37665155576542, 0.0649054821114987, 0.37052654591389, 0.386380397947505, 0.953305578324944, 0.340973854297772, 0.411062113242224, 0.585860836552456, 0.362703745020553, 0.723637734539807, 0.0970077307429165, 0.0887469984591007, 0.0557821169495583, 0.411114145535976, 0.989731068257242, 0.974551122635603, 0.81349238823168, 0.134170187404379, 0.454863896593451, 0.920629747910425, 0.0718251764774323, 0.62178977811709, 0.9208127935417, 0.845446197781712, 0.351291396887973, 0.363953956402838, 0.472790851956233, 0.471397645538673, 0.784782769158483, 0.152135792886838, 0.668838439043611, 0.693388524698094, 0.900453111622483, 0.566805175505579, 0.388827929273248, 0.541417809203267, 0.600088377250358, 0.395733440294862, 0.538613543380052, 0.109132710145786, 0.0875666157808155, 0.274873720249161, 0.901672996580601, 0.991175264818594, 0.552745331544429, 0.665677615441382, 0.798847352853045, 0.0371757235843688, 0.356695061083883, 0.719219537451863, 0.87011916609481, 0.462587790563703, 0.959651909535751, 0.828183559002355, 0.226516331313178, 0.839787097414955, 0.337563505629078, 0.638332171598449, 0.908960658824071, 0.52450591721572, 0.799348689848557, 0.112237949389964, 0.526198521256447, 0.275596257066354, 0.233752127271146, 0.311431813053787, 0.0601114451419562, 0.273297979030758, 0.358695903792977, 0.264826762489974, 0.823142744135112, 0.493538316804916, 0.287708526477218, 0.482223561964929, 0.836446041474119, 0.788618236780167, 0.887438755249605, 0.118805456208065, 0.342685830313712, 0.328108897898346, 0.696744829649106, 0.478702282067388, 0.678328060545027, 0.658263217657804, 0.365692257415503, 0.245191841153428, 0.356771346181631, 0.910475373500958, 0.665887805167586, 0.101979853818193, 0.732033481355757, 0.701131946174428, 0.0326780762989074, 0.157816290156916, 0.0301310850773007, 0.928671072004363, 0.93093270319514, 0.0133064982946962, 0.730972508899868, 0.28448966704309, 0.253443888155743, 0.217966484604403, 0.340165501227602, 0.83435531752184, 0.845472520217299, 0.373164595570415, 0.180807662894949, 0.200393543811515, 0.064784072805196, 0.244436184177175, 0.209201377350837, 0.865914523135871, 0.132928570033982, 0.445834835059941, 0.440243658144027, 0.776094902772456, 0.128740689251572, 0.413219034671783, 0.00737470341846347, 0.129027106799185, 0.504285529488698, 0.156502886675298, 0.289253134047613, 0.380229765083641, 0.818970392923802, 0.148905873764306, 0.981210689526051, 0.303422867553309, 0.504345459863544, 0.928859111387283, 0.364932046504691, 0.821746611502022, 0.54651231970638, 0.663391130510718, 0.18678341479972, 0.639686370035633, 0.81113004591316, 0.980317063396797, 0.17148721171543, 0.887021143687889, 0.971672155661508, 0.09610087540932, 0.528232117649168, 0.219035412650555, 0.18463063868694, 0.797075050184503, 0.18624617648311, 0.536733089247718, 0.835129141574726, 0.621298149228096, 0.773602606961504, 0.438950859475881, 0.596960843075067, 0.335658268770203, 0.134654247201979, 0.437801544554532, 0.474604102550074, 0.744382972130552, 0.663408585358411, 0.0889651449397206, 0.788250726880506, 0.950591837288812, 0.863443164387718, 0.353806459810585, 0.992279505357146, 0.666055221809074, 0.0616062695626169, 0.335183793213218, 0.758372209500521, 0.564091283595189, 0.697035257471725, 0.23092269571498, 0.643045540899038, 0.158920723479241, 0.634148272452876, 0.947597049642354, 0.647944726515561, 0.912229019217193, 0.616160501493141, 0.945166572928429, 0.146978461649269, 0.624227398075163, 0.236259706318378, 0.391387410694733, 0.910806987201795, 0.313424498774111, 0.666922922711819, 0.322931075235829, 0.332710357615724, 0.399841514183208, 0.190625277347863, 0.628739349078387, 0.911890657152981, 0.057751152664423, 0.0744612368289381, 0.23929253895767, 0.387483017286286, 0.96255720523186, 0.359090743586421, 0.574099485063925, 0.0780634719412774, 0.53291982691735, 0.156904214527458, 0.76231803605333, 0.250236185034737, 0.508316231891513, 0.465729085262865, 0.388484843308106, 0.250066727632657, 0.00566646456718445, 0.451842450769618, 0.426615631673485, 0.0831722291186452, 0.531290107173845, 0.588151181582361, 0.0144150573760271, 0.233840567991138, 0.216149910585955, 0.102312278002501, 0.836818124866113, 0.328905911184847, 0.271931905532256, 0.649354382418096, 0.438029748620465, 0.944051263388246, 0.701638192171231, 0.411410686792806, 0.335185671690851, 0.168495317688212, 0.0995507566258311, 0.381141302408651, 0.756156120449305, 0.110433355905116, 0.443068747175857, 0.0172620720695704, 0.74159476114437, 0.604869271861389, 0.969089527847245, 0.937763738678768, 0.879816606873646, 0.116879202891141, 0.968492706073448, 0.838998029241338, 0.159098986303434, 0.371197037631646, 0.122503753285855, 0.763681674608961, 0.933127717813477, 0.145103385439143, 0.584056759718806, 0.0218203272670507, 0.611109957331792, 0.100784548791125, 0.386886026244611, 0.628360261209309, 0.161527439253405, 0.711599403293803, 0.62559473933652, 0.144723201636225, 0.868719586636871, 0.0550054486375302, 0.931937007699162, 0.909042129991576), actual = c(0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"