Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.7840432        0.5199164 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5199164        0.5199164        0.5072818 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.37338843755424, 0.317458307603374, 0.454288656823337, 0.349376000231132, 0.0156010997015983, 0.490312366047874, 0.876000631367788, 0.0934876264072955, 0.407846944406629, 0.562938672024757, 0.0544956284575164, 0.156040129484609, 0.0674605590756983, 0.333870535949245, 0.688516136491671, 0.889684934867546, 0.294230153318495, 0.73900531954132, 0.278812762582675, 0.214620052138343, 0.612914299359545, 0.630624677520245, 0.7613223453518, 0.936711226124316, 0.351609808625653, 0.80837736511603, 0.826834043487906, 0.346964169060811, 0.835877084871754, 0.0945471865125, 0.0650677757803351, 0.912245899206027, 0.126728031318635, 0.0260316070634872, 0.804129466647282, 0.961983235087246, 0.30079234810546, 0.303034708136693, 0.245004480006173, 0.608896300429478, 0.74281343142502, 0.373370306566358, 0.278475601458922, 0.47724023880437, 0.78694430901669, 0.90610023890622, 0.169207307975739, 0.592210251605138, 0.929010591004044, 0.719727599294856, 0.153744377428666, 0.190802867989987, 0.764665528433397, 0.301277491962537, 0.0990886064246297, 0.231842645909637, 0.293239962076768, 0.582883981987834, 0.31740807951428, 0.701240886002779, 0.459922624984756, 0.573823255486786, 0.0413696696050465, 0.41228489857167, 0.700033921049908, 0.909372258232906, 0.260417623911053, 0.796421015402302, 0.63945696875453, 0.257138174725696, 0.796979382168502, 0.49634159472771, 0.00937078986316919, 0.337071590125561, 0.392680854070932, 0.349701133091003, 0.723783834371716, 0.394807674223557, 0.504596157697961, 0.55345451249741, 0.0861767979804426, 0.0784995027352124, 0.421676351223141, 0.490759076550603, 0.213687051320449, 0.0606456974055618, 0.531044630100951, 0.867573140887544, 0.139390212949365, 0.857573590008542, 0.89354043174535, 0.243295367341489, 0.597987620858476, 0.347906212788075, 0.557102604070678, 0.572803880320862, 0.448727614013478, 0.812926970189437, 0.269399964716285, 0.845845312112942, 0.797210627468303, 0.0648811487480998, 0.683738360181451, 0.495382543187588, 0.169473652495071, 0.403415315551683, 0.525968287838623, 0.0374688727315515, 0.876423369860277, 0.324262940790504, 0.69335827883333, 0.118630195967853, 0.817917907843366, 0.970694234827533, 0.494848567526788, 0.165360370418057, 0.0492146334145218, 0.0247278409078717, 0.439488657051697, 0.963079175446182, 0.827084185089916, 0.457681728294119, 0.658519253134727, 0.726465615443885, 0.715031893225387, 0.794502617791295, 0.432665338972583, 0.516229676548392, 0.376745690358803, 0.847813148749992, 0.477060295641422, 0.28157335659489, 0.582074170466512, 0.94723097467795, 0.0307908682152629, 0.300501387333497, 0.668796274345368, 0.130149794509634, 0.337247564923018, 0.232699698535725, 0.354352358495817, 0.0874616694636643, 0.116671689553186, 0.181043198332191, 0.866973334457725, 0.682822219561785, 0.108174307271838, 0.499507651198655, 0.463470558170229, 0.151158907916397, 0.308904331643134, 0.0171670806594193, 0.101306869415566, 0.779652970843017, 0.017077952856198, 0.294635887490585, 0.746873714029789, 0.14495029929094, 0.390497213462368, 0.269174854969606, 0.451864810660481, 0.392331598326564, 0.801060753874481, 0.804175938246772, 0.813901962712407, 0.608050012495369, 0.213544625090435, 0.663901538355276, 0.18456767895259, 0.661453659180552, 0.188534097513184, 0.572937358170748, 0.946499662706628, 0.444317147834226, 0.942180405603722, 0.973316782386974, 0.752529587596655, 0.175484078936279, 0.735744912177324, 0.0902378503233194, 0.933438089909032, 0.558472432661802, 0.792834064923227, 0.532851802185178, 0.287452408578247, 0.309479462215677, 0.0320878638885915, 0.919443565653637, 0.578262908151373, 0.739431000780314, 0.818652774207294, 0.91439812281169, 0.850290548754856, 0.239329048199579, 0.806845882441849, 0.435930997831747, 0.53788172500208, 0.148384720552713, 0.51628439873457, 0.311932912096381, 0.545486605493352, 0.627243243390694, 0.353981944732368, 0.575003838865086, 0.126166897825897, 0.662272387882695, 0.351489507360384, 0.909129329025745, 0.433066952973604, 0.462489015888423, 0.0626955747138709, 0.382095695706084, 0.642504259711131, 0.211555709363893, 0.866019158391282, 0.078791938489303, 0.854335918324068, 0.504654840566218, 0.506512448890135, 0.292644722154364, 0.119698386639357, 0.0193347139284015, 0.610944291343912, 0.969719209242612, 0.40990625647828, 0.971102993236855, 0.54863574099727, 0.199709184002131, 0.524332299130037, 0.0158335722517222, 0.706779900472611, 0.220590704586357, 0.10409614094533, 0.646910508396104, 0.603528286097571, 0.689574090531096, 0.657735615037382, 0.177478206576779, 0.689921835903078, 0.510236276080832, 0.567707373760641, 0.832419305341318, 0.659164790529758, 0.99314134195447, 0.882916203467175, 0.472024240996689, 0.366060408297926, 0.724395926808938, 0.833166642347351, 0.798206531442702, 0.128583687124774, 0.425210563931614, 0.0991878849454224, 0.697543962625787, 0.840082217473537, 0.437542130006477, 0.591369746951386, 0.944119428750128, 0.656624834984541, 0.772605084581301, 0.834692155010998, 0.504410451510921, 0.266844453057274, 0.230588041711599, 0.0752521345857531, 0.0350462747737765, 0.82333937799558, 0.969894289737567, 0.122756326338276, 0.587546054506674, 0.324339398182929, 0.179477129131556, 0.758821381954476, 0.536230814177543, 0.470226553967223, 0.880673011532053, 0.694038503337651, 0.254374515730888, 0.80983197200112, 0.683962742099538, 0.462066816166043, 0.273719046032056, 0.452717508655041, 0.900775087997317, 0.359933971660212, 0.689639722695574, 0.984824101673439, 0.475789737887681, 0.931919750059023, 0.215616028988734, 0.298713364172727, 0.915550102712587, 0.664698414504528, 0.606666773557663, 0.0242494912818074, 0.92917305091396, 0.982436767080799, 0.131504555232823, 0.511408166028559, 0.694883917924017, 0.968442403711379, 0.811597047140822, 0.219110340811312, 0.424394515575841, 0.291995948180556, 0.479253610596061, 0.266256748931482, 0.352991274325177, 0.485906790243462, 0.693896539509296, 0.612415040610358, 0.767435744870454, 0.596444050315768, 0.403807955561206, 0.562900305259973, 0.706789529649541, 0.793446691473946, 0.229143305681646, 0.840523477178067, 0.398191357729957, 0.508272115141153, 0.897722903871909, 0.726578630041331, 0.673409123672172, 0.150171785615385, 0.832862091017887, 0.253777127480134, 0.425752847455442, 0.431718863081187, 0.365893530193716, 0.865966683253646, 0.329601056408137, 0.75101000093855, 0.39329211669974, 0.887529341503978, 0.836621284252033, 0.659125656355172, 0.0744700329378247, 0.495564093580469, 0.143114859238267, 0.195318590616807, 0.385071660857648, 0.618979271035641, 0.0601181911770254, 0.130780570674688, 0.00693102949298918, 0.0968380672857165, 0.539589481195435, 0.950308130821213, 0.611104811541736, 0.283631415106356, 0.247236499097198, 0.805897323647514, 0.462114828405902, 0.0231125201098621, 0.40549327339977, 0.592524005565792, 0.176836888771504, 0.461771696107462, 0.398280475754291, 0.906102043110877, 0.315054325852543, 0.420423260424286, 0.267396976239979, 0.554302841657773, 0.170647443737835, 0.848023061873391, 0.405852949479595, 0.746969460276887, 0.834284447832033, 0.462745051132515, 0.185597050236538, 0.955224449280649, 0.471158991102129, 0.640819407533854, 0.234344203956425, 0.897150822449476, 0.199046315858141, 0.4889048805926, 0.904331343714148, 0.0939937736839056, 0.0800958930049092, 0.292642480460927, 0.637636960251257, 0.95666822604835, 0.588603608077392, 0.27455474413, 0.275915214093402, 0.888788733864203, 0.894181350711733, 0.00182435638271272, 0.165042251348495, 0.648736592382193, 0.952916910871863, 0.297199941240251, 0.658007940510288, 0.971174486214295, 0.0241827233694494, 0.163489437196404, 0.191447082906961, 0.11624477384612, 0.877893386874348, 0.584904148243368, 0.217839839635417, 0.146383628016338, 0.935723234666511, 0.079848489491269, 0.248079363489524, 0.403153663268313, 0.0431833053007722, 0.978812180925161, 0.802367764292285, 0.242134645814076, 0.263091639149934, 0.340242257108912, 0.534448784310371, 0.249246925581247, 0.0307287673931569, 0.574258212931454, 0.973054276546463, 0.703052847180516, 0.89146409323439, 0.778100416064262, 0.638079025782645, 0.815361813874915, 0.0209947489202023, 0.779315516585484, 0.901951040141284, 0.394277081824839, 0.974258661270142, 0.49310919479467, 0.519353937357664, 0.229268671013415, 0.0762383954133838, 0.420332387089729, 0.212591128889471, 0.220366186695173, 0.920746720628813, 0.558654703898355, 0.382164251990616, 0.938551477156579, 0.520998494932428, 0.0576889766380191, 0.674764488358051, 0.311937619699165, 0.0761176315136254, 0.0500603497494012, 0.629653526470065, 0.806562457466498, 0.299929635133594, 0.545900581404567, 0.484279888682067, 0.774842349346727, 0.963752876035869, 0.208357856608927, 0.451201519928873, 0.213580484502017, 0.223859746241942, 0.697003395529464, 0.437094248831272, 0.629088405752555, 0.891214789822698, 0.0990780566353351, 0.00195362768135965, 0.371172617888078, 0.583178607514128, 0.328532523242757, 0.470764888217673, 0.679400059394538, 0.739325446775183, 0.821566191269085, 0.610595287755132, 0.895011758664623, 0.303346210625023, 0.103402649052441, 0.618487958563492, 0.0489756842143834, 0.333678814116865, 0.662556130439043, 0.256747521460056, 0.337896182434633, 0.837097768206149, 0.398969619302079, 0.616475032642484, 0.583429533522576, 0.035139980725944, 0.231633946066722, 0.68111160164699, 0.620407991111279, 0.703488511499017, 0.639432806288823, 0.396290187491104, 0.34658225462772, 0.53092581499368, 0.308857229305431, 0.097631732467562, 0.754096296150237, 0.084414237877354, 0.184874223545194, 0.759608196560293, 0.623932752292603, 0.421999559272081, 0.071027202764526, 0.648622326785699, 0.979836928192526, 0.59183765691705, 0.410309152677655, 0.353931727586314, 0.235679391073063, 0.426744029158726, 0.662629632046446, 0.0366519733797759, 0.136384700424969, 0.718483501113951, 0.566072646295652, 0.998192306607962, 0.635803324636072, 0.324528693221509, 0.211495090741664, 0.364764435682446, 0.973870288114995, 0.423422791529447, 0.0525089758448303, 0.0838721210602671, 0.0842099264264107, 0.254522101487964, 0.705826547695324, 0.742949296487495, 0.911366566317156, 0.384655645117164, 0.00540580763481557, 0.00981710385531187, 0.781709843082353, 0.0256824302487075, 0.450308305444196, 0.316342322621495, 0.77700657164678, 0.374231316382065, 0.461404347792268, 0.760315202875063, 0.60721609252505, 0.173533270368353, 0.644905522931367, 0.33580171992071, 0.446151118958369, 0.0388317778706551, 0.683454306796193, 0.113201067782938, 0.751937387976795, 0.375353514449671, 0.143252808833495, 0.394111961359158, 0.313380581093952, 0.432394968112931, 0.302982961060479, 0.486573209986091, 0.184818115318194, 0.0557784126140177, 0.124452884774655, 0.864337200764567, 0.797424490796402, 0.110953667899594, 0.786690382519737, 0.406769891735166, 0.0378407461103052, 0.787173492833972, 0.165270475437865, 0.61642022896558, 0.735921908169985, 0.0249455505982041, 0.709143500309438, 0.466497047571465, 0.0686470221262425, 0.15820446331054, 0.0596450574230403, 0.274990974692628, 0.976811161031947, 0.380441016051918, 0.883400406688452, 0.81954929116182, 0.248923386214301, 0.611521228449419, 0.0892929995898157, 0.148059126455337, 0.27148616570048, 0.114029733464122, 0.803891085786745, 0.196442143525928, 0.174525613198057, 0.346220073290169, 0.0941134956665337, 0.440346189308912, 0.310518379788846, 0.893932395149022, 0.841861535795033, 0.17128110001795, 0.277188017498702, 0.559531866107136, 0.506146010709926, 0.885980732738972, 0.481396253220737, 0.516604070551693, 0.928992525674403, 0.652269429061562, 0.639245374128222, 0.927633716724813, 0.114438576158136, 0.901139360386878, 0.173658326501027, 0.712673099711537, 0.187319424469024, 0.429554360453039, 0.290004410082474, 0.968110839836299, 0.0412490624003112, 0.858428540173918, 0.386836977442726, 0.00118864444084466, 0.786950232228264, 0.686544945696369, 0.0846280385740101, 0.850875620963052, 0.93617343227379, 0.983895983546972, 0.978800448123366, 0.18180861719884, 0.888260685140267, 0.317577047506347, 0.986257331212983, 0.0150739275850356, 0.0182101354002953, 0.74107142817229, 0.926107147475705, 0.828014655271545, 0.57781134895049, 0.487619245424867, 0.649762144312263, 0.0282033514231443, 0.161940030287951, 0.28014822723344, 0.606952907051891, 0.194000612711534, 0.274377383757383, 0.630680202273652, 0.451511914143339, 0.826368889771402, 0.679630402009934, 0.982143254252151, 0.467361003160477, 0.455439282814041, 0.96257973369211, 0.270568115403876, 0.182574230711907, 0.100082964403555, 0.273638373240829, 0.846245266031474, 0.56047041201964, 0.551375933922827, 0.0260543180629611, 0.764659473672509, 0.446670929668471, 0.896644672611728, 0.432253250619397, 0.303721311036497, 0.477873888565227, 0.379956942750141, 0.720100004691631, 0.8792655011639, 0.449817333370447, 0.0887395786121488, 0.987561739282683, 0.428984270896763, 0.526685903547332, 0.693732404150069, 0.0700230696238577, 0.769997680326924, 0.187888492131606, 0.582836200483143, 0.909681257326156, 0.681716558756307, 0.965116133214906, 0.532800636487082, 0.509612439433113, 0.848246314795688, 0.916577513562515, 0.915710132103413, 0.304497059900314, 0.777934823883697, 0.570645658997819, 0.809350739931688, 0.0296370175201446, 0.284607126843184, 0.415923124179244, 0.314014349831268, 0.691962655400857, 0.17136619403027, 0.508465475402772, 0.499410361051559, 0.436759883537889, 0.481171766296029, 0.896608940325677, 0.489347661379725, 0.213446445763111, 0.624937003012747, 0.621363213518634, 0.424521893030033, 0.81419203709811, 0.188637910410762, 0.952391546219587, 0.301052134949714, 0.0987433702684939, 0.418264639098197, 0.0240213363431394, 0.0171250104904175, 0.698418715037405, 0.188303010072559, 0.763188918819651, 0.594416120555252, 0.924251812277362, 0.542033424600959, 0.575337875634432, 0.178326737834141, 0.862467688973993, 0.300976366270334, 0.220707711996511, 0.17640587175265, 0.989304284099489, 0.792013076366857, 0.97254593251273, 0.205535635584965, 0.475440284935758, 0.236263704486191, 0.907554140780121, 0.191321752034128, 0.478307011537254, 0.408090106677264, 0.534694333793595, 0.917266476899385, 0.967214732663706, 0.0150138018652797, 0.645923711359501, 0.238785519497469, 0.527278337627649, 0.287536387564614, 0.676728429272771, 0.644324100576341, 0.954409854253754, 0.421952073927969, 0.946362691698596, 0.0629471559077501, 0.916550362017006, 0.841012401273474, 0.195974099915475, 0.181657935259864, 0.0545179941691458, 0.325001718709245, 0.548631608486176, 0.675818520598114, 0.439410123974085, 0.284220694564283, 0.697081679245457, 0.293815984390676, 0.020665131509304, 0.089558515464887, 0.770332037005574, 0.460617212112993, 0.275005831615999, 0.760834793327376, 0.71501258155331, 0.4267664977815, 0.444060307228938, 0.811522726435214, 0.287015529116616, 0.942193150985986, 0.84243939560838, 0.218918161233887, 0.92732156557031, 0.948901102412492, 0.619515079073608, 0.539688289398327, 0.518486476270482, 0.00951250293292105, 0.556502916384488, 0.856706099817529, 0.393228258937597, 0.893633554689586, 0.992997838184237, 0.739797888789326, 0.763685530051589, 0.487801919458434, 0.488892741966993, 0.200452574994415, 0.563503166195005, 0.63705897401087, 0.548096300801262, 0.199337923200801, 0.239114725729451, 0.739086313173175, 0.875130561413243, 0.23223173385486, 0.742536418838426, 0.570604619802907, 0.86938538053073, 0.732098604785278, 0.617553250864148, 0.312366199679673, 0.0631107767112553, 0.670777323190123, 0.245279486989602, 0.934124755207449, 0.70837571262382, 0.680609388044104, 0.586756102507934, 0.371577775804326, 0.611172967357561, 0.290526477852836, 0.429572482127696, 0.823359775589779, 0.10914100962691, 0.958804200170562, 0.901184651535004, 0.0663783079944551, 0.966104328399524, 0.509330191183835, 0.745843078009784, 0.793701054761186, 0.160348650533706, 0.802240854827687, 0.273625941248611, 0.721474934834987, 0.194673632737249, 0.834137007594109, 0.0139933705795556, 0.892360709607601, 0.952399424044415, 0.392560041043907, 0.379317864077166, 0.456588148139417, 0.608524554874748, 0.581856135744601, 0.882127515971661, 0.985837610205635, 0.105922757415101, 0.919616509228945, 0.86361598642543, 0.127769846469164, 0.214024529792368, 0.456168966833502, 0.554233183385804, 0.766004021046683, 0.941646106541157, 0.309746378334239, 0.563236100133508, 0.364045137306675, 0.335834808880463, 0.0414745253510773, 0.989956302335486, 0.541765288216993, 0.867194572929293, 0.264328446937725, 0.197330225491896, 0.268319954629987, 0.465176094323397, 0.800721974577755, 0.117104669334367, 0.948065485805273, 0.37255704565905, 0.853980620391667, 0.793204847490415, 0.847778104012832, 0.93803877918981, 0.155291306087747, 0.445972758578137, 0.942093089688569, 0.407075657509267, 0.320084093371406, 0.213196024065837, 0.518673372920603, 0.53911486431025, 0.768811824964359, 0.721137163927779, 0.794288252713159, 0.868335957638919, 0.0520121825393289, 0.66650493000634, 0.430475041270256, 0.765728622907773, 0.579996206564829, 0.258532349020243, 0.217162725050002, 0.698783650295809, 0.429300353163853, 0.468787813326344, 0.437004016712308, 0.663241130299866, 0.612977919867262, 0.649555538548157, 0.695622120518237, 0.712894877186045, 0.857898217625916, 0.422698809765279, 0.629158186260611, 0.755191701231524, 0.960725755896419, 0.317904788302258, 0.870863449759781, 0.928696242161095, 0.189595878124237, 0.498189074220136, 0.377027055481449, 0.541159401880577, 0.827991730067879, 0.878175301244482, 0.419784651137888, 0.37157109263353, 0.806291211629286, 0.200844127684832, 0.843983647879213, 0.73470424185507, 0.859528009081259, 0.0913545368239284, 0.607616616878659, 0.784759766887873, 0.857663688017055, 0.546348738251254, 0.751050742808729, 0.388454620027915, 0.0635642525739968, 0.572987624211237, 0.456140250666067, 0.845273054903373, 0.668771194759756, 0.0171072559896857, 0.514700765255839, 0.691789569798857, 0.40156162972562, 0.850202774628997, 0.915387769462541, 0.928067909786478, 0.43826285074465, 0.183319233590737, 0.0914124615956098, 0.616714116651565, 0.023802237585187, 0.448641622439027, 0.979832907905802, 0.07054522796534, 0.150052214041352, 0.859021253651008, 0.652233912143856, 0.0672325575724244, 0.127406756626442, 0.757125604199246, 0.412028721766546, 0.365515627432615, 0.180749194929376, 0.343218306079507, 0.691552694421262, 0.74129212833941, 0.39721533539705, 0.771090906811878, 0.955491173779592, 0.953224623575807, 0.379110928624868, 0.773253067629412, 0.665778709342703, 0.538239305606112, 0.749856091104448, 0.770697336178273, 0.616517252521589, 0.276827732566744, 0.313515058718622, 0.679028506623581, 0.646399087505415, 0.494885684922338, 0.484842322999611, 0.301021076738834, 0.674918330507353, 0.68281869427301, 0.126839306671172, 0.886124315438792, 0.109074807958677, 0.134926069527864, 0.973770717391744, 0.679720471613109, 0.694720439612865, 0.946519654709846, 0.982684979680926, 0.384293965762481, 0.931890457635745, 0.221561067504808, 0.287717015482485, 0.737573533086106, 0.460143852513283, 0.73463934683241, 0.747401681961492, 0.281818786635995, 0.209574803011492, 0.322489977348596, 0.19549736334011), actual = c(1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.724010317353532, 0.983417709358037, 0.120225903345272, 0.79517281264998, 0.810889564221725, 0.103835220914334, 0.00738015328533947, 0.0649029994383454, 0.144449303392321, 0.385390377603471, 0.394329039612785, 0.888157562818378, 0.0244065404403955, 0.617715688422322, 0.443051984766498, 0.944112833589315, 0.697078432422131, 0.167025912553072, 0.294478641822934, 0.973761801142246, 0.787233448820189, 0.341667452361435, 0.257209965726361, 0.617050720844418, 0.648304706905037, 0.444151711184531, 0.624010419938713, 0.983202527277172, 0.558321175631136, 0.852306057000533, 0.195740245981142, 0.572490748949349, 0.908584362827241, 0.534026553155854, 0.749264260521159, 0.789901366923004, 0.184188035083935, 0.0550933626946062, 0.048024064861238, 0.276389508275315, 0.200105259893462, 0.622013072716072, 0.493538386654109, 0.67116555175744, 0.0457559730857611, 0.970558613771573, 0.442108173156157, 0.0131988141220063, 0.38515701610595, 0.0970000545494258, 0.37406321358867, 0.869866769295186, 0.67916972306557, 0.0597502693999559, 0.600000754930079, 0.450279283337295, 0.722846287535504, 0.670848719310015, 0.10351763968356, 0.421401008265093, 0.210655277362093, 0.485897625330836, 0.0395889452192932, 0.806172380922362, 0.287200199672952, 0.882560751400888, 0.0808480461128056, 0.458498451160267, 0.734301738673821, 0.59789580036886, 0.15287738153711, 0.313556088600308, 0.707852845080197, 0.419266897020862, 0.324992552632466, 0.635304261231795, 0.688358757644892, 0.0178799494169652, 0.727933395653963, 0.951920855557546, 0.591481718001887, 0.876720000291243, 0.241130070062354, 0.135526253841817, 0.200807509478182, 0.352466549025849, 0.503388402285054, 0.954907777020708, 0.374872053274885, 0.555690391920507, 0.812573396367952, 0.119838911341503, 0.197670233668759, 0.607460920698941, 0.642869692761451, 0.00415348191745579, 0.277287666220218, 0.301842644345015, 0.672216125531122, 0.896206858800724, 0.0681024480145425, 0.659486510790884, 0.318308760412037, 0.277030267287046, 0.835587340872735, 0.747267783153802, 0.194867887534201, 0.792591166915372, 0.968678148463368, 0.11748769483529, 0.366182035300881, 0.74678948149085, 0.243018909590319, 0.147478280588984, 0.544008398195729, 0.111506859306246, 0.0794921643100679, 0.21770854620263, 0.1732740229927, 0.903255283366889, 0.32971542654559, 0.615120916394517, 0.207086323527619, 0.242124411975965, 0.822077176533639, 0.389374603051692, 0.501426739851013, 0.112704997882247, 0.970307245850563, 0.215887624071911, 0.737553742947057, 0.071137769613415, 0.224225328071043, 0.839046685956419, 0.0875353014562279, 0.472884693648666, 0.156055369181558, 0.439669524552301, 0.217316318536177, 0.568172539118677, 0.712961461395025, 0.977158571127802, 0.941058934666216, 0.701488794991747, 0.638380363816395, 0.220294686499983, 0.570245493669063, 0.945709690218791, 0.293503906577826, 0.1160262695048, 0.631306400056928, 0.674315535230562, 0.599742077523842, 0.384743654401973, 0.811707767192274, 0.148015597602352, 0.740263669053093, 0.478401810396463, 0.000485791591927409, 0.455605052178726, 0.966852127108723, 0.551014745142311, 0.670328575419262, 0.0353274152148515, 0.462729941587895, 0.448864101199433, 0.760968622984365, 0.938076080521569, 0.6787410590332, 0.942320274887607, 0.149595242459327, 0.123332482529804, 0.444460237398744, 0.372919878922403, 0.452739434782416, 0.521615322912112, 0.667609310010448, 0.456591885304078, 0.404626357369125, 0.192769014276564, 0.487239704234526, 0.673568616854027, 0.333633615868166, 0.781899292487651, 0.607570900116116, 0.869801341090351, 0.891441590618342, 0.741572985192761, 0.85919012944214, 0.879323227098212, 0.606799885863438, 0.66995137417689, 0.787248694803566, 0.714458245318383, 0.431762730702758, 0.20775031321682, 0.907861879095435, 0.428530778037384, 0.982151476666331, 0.0397446968127042, 0.953690685564652, 0.491040796274319, 0.0604436357971281, 0.103074284037575, 0.284844706766307, 0.488568681757897, 0.711119830142707, 0.0192998822312802, 0.331746024312451, 0.015514463884756, 0.0528631687629968, 0.859948216471821, 0.992821066873148, 0.0289037197362632, 0.514250296633691, 0.877349894028157, 0.911822651745752, 0.861521536484361, 0.0199271419551224, 0.736723565263674, 0.652182966470718, 0.580900605069473, 0.350769714219496, 0.729436537250876, 0.645857680821791, 0.269477083813399, 0.0190055500715971, 0.647165002301335, 0.894476626068354, 0.493316253181547, 0.0803635471966118, 0.764064396265894, 0.889214815571904, 0.95714297099039, 0.14466711319983, 0.871607309905812, 0.64407489541918, 0.215169485658407, 0.841114563634619, 0.835914425551891, 0.583197111729532, 0.887695949524641, 0.683321515098214, 0.581732351565734, 0.111350332153961, 0.494584781816229, 0.703639816958457, 0.439027207437903, 0.449147711275145, 0.536688320804387, 0.918827019864693, 0.963005532044917, 0.402019141241908, 0.332934293197468, 0.0988406892865896, 0.0636278896126896, 0.0304742262233049, 0.797569807153195, 0.52570586046204, 0.126623564399779, 0.632991244085133, 0.560436236439273, 0.0951208036858588, 0.0267756427638233, 0.95167745067738, 0.505085218464956, 0.979989168467, 0.383407764136791, 0.875973861664534, 0.690747656626627, 0.373839269625023, 0.443761944072321, 0.333230215823278, 0.609934720909223, 0.344491451047361, 0.118922652443871, 0.0742290550842881, 0.313987248577178, 0.855406007263809, 0.494006597902626, 0.00644345371983945, 0.347046536859125, 0.428991688182577, 0.766541321529076, 0.988583728205413, 0.615975425112993, 0.450998142361641, 0.308280434925109, 0.76373926596716, 0.716384019469842, 0.418181241257116, 0.111273527611047, 0.682968377135694, 0.78600972192362, 0.263509994139895, 0.194045040989295, 0.107871661894023, 0.34777031163685, 0.965969480806962, 0.968720271019265, 0.724441159516573, 0.850073736859486, 0.591147835366428, 0.229052315000445, 0.0462067453190684, 0.382219053572044, 0.559560116380453, 0.976468258304521, 0.521501942304894, 0.804951050551608, 0.315633712802082, 0.234405538067222, 0.0520271882414818, 0.424825107911602, 0.724113750271499, 0.698923899326473, 0.288571564015001, 0.265338745201007, 0.0977976168505847, 0.130984402727336, 0.792191047221422, 0.352092328015715, 0.366639350773767, 0.467624750453979, 0.363921323558316, 0.678543259389699, 0.907868977170438, 0.251227099448442, 0.357318901689723, 0.909340751823038, 0.114265406271443, 0.874310368904844, 0.875476014101878, 0.0248183791991323, 0.979467683937401, 0.154860365204513, 0.139333937317133, 0.0706230939831585, 0.0247300253249705, 0.71085229027085, 0.271889970405027, 0.631822074996307, 0.408881091279909, 0.0183267567772418, 0.330886213807389, 0.859925658674911, 0.73181524220854, 0.302970023592934, 0.537989055737853, 0.257585607003421, 0.27099318173714, 0.688359605614096, 0.552820737008005, 0.6314001979772, 0.509027077350765, 0.574709748849273, 0.930849133990705, 0.490729677258059, 0.807746869511902, 0.918517265003175, 0.606431433698162, 0.890262112021446, 0.829875137656927, 0.532307801768184, 0.3889060462825, 0.710649328073487, 0.259344226447865, 0.536448531085625, 0.760711198672652, 0.48717027483508, 0.973710191668943, 0.0294226305559278, 0.323021934833378, 0.489582523703575, 0.999294255161658, 0.5584856136702, 0.893323746742681, 0.806062061805278, 0.777916622813791, 0.132686204044148, 0.644473547581583, 0.868591844802722, 0.205071174772456, 0.0462184366770089, 0.863993972307071, 0.383230890147388, 0.813649496296421, 0.810310840606689, 0.189794628880918, 0.936213015113026, 0.666184693342075, 0.845417537959293, 0.924755634739995, 0.502896170364693, 0.359680292895064, 0.768014712026343, 0.716931227594614, 0.403287951601669, 0.940062533365563, 0.609880089294165, 0.0396623234264553, 0.555121469311416, 0.0100275913719088, 0.996479456312954, 0.728659430053085, 0.7375435649883, 0.253200828330591, 0.232048061676323, 0.303051610710099, 0.421330019831657, 0.228041258640587, 0.772543538361788, 0.040398585377261, 0.775828085374087, 0.372728157555684, 0.603488134220243, 0.630468415329233, 0.257630777312443, 0.647991016507149, 0.367154946550727, 0.195241857785732, 0.0433549946174026, 0.561145955231041, 0.547975104069337, 0.473733997903764, 0.27973105898127, 0.724482031073421, 0.344678778434172, 0.849768112879246, 0.0329936305060983, 0.789791538380086, 0.146306874696165, 0.373293580021709, 0.24040735187009, 0.260096273384988, 0.749495420837775, 0.820944293402135, 0.495690215379, 0.951510168146342, 0.170738176908344, 0.0619329276960343, 0.439778275089338, 0.0102913130540401, 0.078229685081169, 0.925817270064726, 0.746463452698663, 0.165872795972973, 0.887231530155987, 0.686228396603838, 0.337634433060884, 0.892696525668725, 0.685917251743376, 0.0672555109485984, 0.350284578977153, 0.321705179987475, 0.223276819335297, 0.991285299183801, 0.379466057987884, 0.135514096589759, 0.515524270245805, 0.0572407366707921, 0.0778340680990368, 0.262174187926576, 0.929091294761747, 0.914936227491125, 0.929750308860093, 0.938748195301741, 0.763008264591917, 0.418301807483658, 0.118473780807108, 0.262289397418499, 0.287212409079075, 0.82502783765085, 0.758805563906208, 0.926533498102799, 0.741481594508514, 0.497823727084324, 0.661607925547287, 0.330504877958447, 0.53908864618279, 0.900012595346197, 0.781595714157447, 0.341104650171474, 0.3756648756098, 0.649860222591087, 0.404574765590951, 0.211239349795505, 0.55532717355527, 0.844501530518755, 0.00277647818438709, 0.108481161063537, 0.469066893216223, 0.614298548316583, 0.274378356290981, 0.308149423915893, 0.903299655066803, 0.547281966544688, 0.272219793405384, 0.145376833621413, 0.952488442184404, 0.949397129472345, 0.656296937959269, 0.358141641132534, 0.654320388333872, 0.131095130229369, 0.240342821227387, 0.727687784470618, 0.437340927310288, 0.791485938243568, 0.0855906447395682, 0.578317576553673, 0.400587774580345, 0.847849957179278, 0.934757591923699, 0.0135366993490607, 0.207668232033029, 0.519981837365776, 0.0543162382673472, 0.142139340983704, 0.0555428136140108, 0.625941592268646, 0.104897536803037, 0.733580907806754, 0.602650864748284, 0.580171716399491, 0.0879996747244149, 0.506750415777788, 0.650985655607656, 0.909172876970842, 0.640720706898719, 0.934139501303434, 0.860274776816368, 0.397012040251866, 0.820930920774117, 0.39249206148088, 0.15500727458857, 0.69775902479887, 0.95634049596265, 0.513481647707522, 0.0444168939720839, 0.462031615432352, 0.0825990242883563, 0.344497633632272, 0.00437552994117141, 0.273972243070602, 0.497471456881613, 0.142687289742753, 0.702138889115304, 0.723066804697737, 0.0490605251397938, 0.922861275263131, 0.0149196130223572, 0.0976231314707547, 0.377600951818749, 0.634509949013591, 0.744993937667459, 0.456274110591039, 0.0474774616304785, 0.778899823315442, 0.0820979156997055, 0.346507674083114, 0.635292421793565, 0.114356925943866, 0.770711149089038, 0.22920593409799, 0.885871779406443, 0.13943519955501, 0.825960508547723, 0.954878836404532, 0.576021696208045, 0.492417351342738, 0.690085905138403, 0.139647506177425, 0.274339283583686, 0.123143635690212, 0.397177482256666, 0.568934099283069, 0.435811874456704, 0.357060310896486, 0.941690654028207, 0.566405386431143, 0.820886386325583, 0.546526932390407, 0.976886070566252, 0.256205387646332, 0.318034183699638, 0.333117449423298, 0.259998255874962, 0.205272290622815, 0.309344408567995, 0.497204181971028, 0.753442608285695, 0.966629842296243, 0.0900738558266312, 0.415785228367895, 0.0015759349334985, 0.0218998065683991, 0.26456207735464, 0.32382620847784, 0.530986862257123, 0.0193529848475009, 0.700888219056651, 0.403739469125867, 0.656195633579046, 0.661210871534422, 0.728029486257583, 0.745032472768798, 0.0326062613166869, 0.639910843456164, 0.43083115410991, 0.938739139121026, 0.19442205876112, 0.0785646070726216, 0.874359035398811, 0.32004687259905, 0.972579706460238, 0.51555467932485, 0.400211042258888, 0.179643637035042, 0.168470012489706, 0.26812134264037, 0.719783298438415, 0.490559815894812, 0.1097071247641, 0.349150426452979, 0.64463006798178, 0.69848891813308, 0.0765316276811063, 0.295434925472364, 0.903445684351027, 0.647254703100771, 0.692798575386405, 0.283133911900222, 0.633081327890977, 0.546554838772863, 0.0211644645314664, 0.585830628871918, 0.916426390875131, 0.739181317621842, 0.991119699785486, 0.40026502776891, 0.683376396074891, 0.663086594780907, 0.487533784471452, 0.0652422381099313, 0.981240696273744, 0.489420316880569, 0.752955075120553, 0.0242613833397627, 0.69326999085024, 0.694308846024796, 0.518080652458593, 0.643482193583623, 0.780629520537332, 0.992208807729185, 0.245028584962711, 0.237232879502699, 0.657245961250737, 0.728643990587443, 0.405474978499115, 0.814818918704987, 0.634452127851546, 0.341215304099023, 0.941852376563475, 0.27206820063293, 0.912628581514582, 0.562923468649387, 0.651348387822509, 0.623757858760655, 0.51337505178526, 0.183302754536271, 0.558848335407674, 0.64905312191695, 0.0238785562105477, 0.614515308523551, 0.170128026977181, 0.559051167918369, 0.656785597791895, 0.975012683076784, 0.150350339710712, 0.577537538250908, 0.260223315330222, 0.683372823754326, 0.742569942958653, 0.685403906041756, 0.481083513004705, 0.633457551943138, 0.995882376097143, 0.529295639833435, 0.088225610088557, 0.321530020097271, 0.107482932042331, 0.534198218258098, 0.621901209931821, 0.707735433243215, 0.481392646208405, 0.961399205494672, 0.715775065124035, 0.46784329530783, 0.137793208938092, 0.188951903488487, 0.817976867547259, 0.169241260271519, 0.179037763504311, 0.933375751366839, 0.268501603510231, 0.506880631670356, 0.337001852691174, 0.165993332164362, 0.559938806807622, 0.688210861757398, 0.959565918426961, 0.439849372720346, 0.831970270723104, 0.697905207751319, 0.16464101197198, 0.987852304242551, 0.671603103866801, 0.0223468805197626, 0.330480830278248, 0.779108650516719, 0.0202752400655299, 0.445299265440553, 0.985098442295566, 0.774926868034527, 0.253100302070379, 0.779913339531049, 0.0192887973971665, 0.663015856174752, 0.104910291498527, 0.220145541476086, 0.766364651732147, 0.745535121764988, 0.351588982157409, 0.281845919089392, 0.914454412879422, 0.584659822285175, 0.831476392224431, 0.171120957005769, 0.000156495487317443, 0.0647944156080484, 0.350628427229822, 0.875244445865974, 0.462906170170754, 0.23978553339839, 0.410750068956986, 0.198335797991604, 0.763982103671879, 0.441623315680772, 0.451969061046839, 0.917497367830947, 0.484881665790454, 0.589966930449009, 0.774535147706047, 0.699085947824642, 0.622391507728025, 0.619789242278785, 0.095394893316552, 0.129587335279211, 0.0138914075214416, 0.285236371681094, 0.393202228704467, 0.490203045541421, 0.677545068087056, 0.647292476845905, 0.714319093385711, 0.459472035290673, 0.922489503631368, 0.105045146308839, 0.0187003368046135, 0.0645401803776622, 0.53346641198732, 0.496501092799008, 0.123013741569594, 0.0307061516214162, 0.559346554102376, 0.567934691905975, 0.975384262390435, 0.801985493395478, 0.185650499071926, 0.438063315814361, 0.198699729982764, 0.830140383914113, 0.12859924393706, 0.570858147460967, 0.613116691820323, 0.634287869790569, 0.892299562226981, 0.0328888418152928, 0.755126845324412, 0.487229786580428, 0.176360681653023, 0.39577332790941, 0.0496861890424043, 0.886010234709829, 0.714244416216388, 0.696533032925799, 0.0651818355545402, 0.297505938215181, 0.434058386832476, 0.0768716942984611, 0.169902148190886, 0.647497423458844, 0.744040103396401, 0.151561614358798, 0.420012667309493, 0.051609184127301, 0.901016008807346, 0.439354434842244, 0.318568987073377, 0.388875580159947, 0.77791919070296, 0.221902290824801, 0.798393317498267, 0.00866391067393124, 0.688618673942983, 0.269789877347648, 0.323690102435648, 0.711333528161049, 0.455139457248151, 0.840956021333113, 0.42659874772653, 0.0496177584864199, 0.768493076087907, 0.804549914319068, 0.19865769916214, 0.996127142570913, 0.417894002981484, 0.763695030473173, 0.978366431780159, 0.812154985032976, 0.53440934070386, 0.815202994272113, 0.0744501210283488, 0.574314811266959, 0.449540358502418, 0.454038641415536, 0.0519897236954421, 0.598429834702983, 0.692486113402992, 0.446504439227283, 0.12302304781042, 0.398568795761093, 0.801168447360396, 0.998799905646592, 0.957478804979473, 0.965023605851457, 0.779449114110321, 0.582972392439842, 0.711289588361979, 0.949772556545213, 0.265520060202107, 0.0319147587288171, 0.274734633974731, 0.0166046370286494, 0.554351861821488, 0.148511258885264, 0.835385154932737, 0.0980584118515253, 0.172401479678228, 0.703148437663913, 0.996445331955329, 0.118704737629741, 0.452960898634046, 0.943927359767258, 0.857810989022255, 0.228973997058347, 0.258838221430779, 0.830861683934927, 0.690380678744987, 0.641430914867669, 0.780959161696956, 0.94304173393175, 0.975582865299657, 0.974430470494553, 0.864126045256853, 0.814729955047369, 0.829509505769238, 0.370578659931198, 0.820722775068134, 0.677079993765801, 0.340598300332204, 0.0272198230959475, 0.846900804899633, 0.509245988214388, 0.737315469654277, 0.47472793306224, 0.553522446891293, 0.104988964274526, 0.392656178213656, 0.803169643273577, 0.519308059243485, 0.840095816412941, 0.385979055427015, 0.0549660900142044, 0.856370580848306, 0.238975938642398, 0.182875714963302, 0.703058824408799, 0.102004958782345, 0.871232809731737, 0.541571981739253, 0.317750252783298, 0.207452929811552, 0.984053979627788, 0.0710267811082304, 0.976497033145279, 0.78756886231713, 0.56071726186201, 0.296286401338875, 0.2920636550989, 0.0663845171220601, 0.282391565153375, 0.282030225265771, 0.888132531195879, 0.23073896416463, 0.286297870567068, 0.396809431258589, 0.857573056826368, 0.230016798013821, 0.833453596802428, 0.583793288795277, 0.755776143399999, 0.148964517284185, 0.567704681307077, 0.603186291176826, 0.611085961339995, 0.404801630647853, 0.461758359801024, 0.916478532133624, 0.285987916635349, 0.091146458638832, 0.770078843459487, 0.713353465078399, 0.610678941244259, 0.540598610416055, 0.0587602143641561, 0.145742572145537, 0.942501567071304, 0.968114315532148, 0.144273756071925, 0.170284663094208, 0.792338375002146, 0.108999195508659, 0.481158086564392, 0.633523308904842, 0.813014726620167, 0.800750928698108, 0.813781416043639, 0.154286901466548, 0.496523214271292, 0.266460966784507, 0.322519581066445, 0.987681195372716, 0.348289855523035, 0.382385730277747, 0.8095775146503, 0.0959224766120315, 0.946609187405556, 0.153434385778382, 0.710058526368812, 0.747864931356162, 0.0639929401222616, 0.265143998200074, 0.0716796442866325, 0.795052019646391, 0.151902098208666, 0.678605498746037, 0.458837955724448, 0.475708466023207, 0.935310336062685, 0.865059317788109, 0.792995896190405, 0.361813226947561, 0.479052985319868, 0.969135987339541, 0.442384708672762, 0.50104652158916, 0.391246293205768, 0.397988268174231, 0.302522982470691, 0.08887132932432, 0.537722205277532, 0.371376286493614, 0.108665900304914, 0.0553497395012528, 0.67583072069101, 0.370361680630594, 0.36237579328008, 0.762447607703507, 0.417064801324159, 0.219282985664904, 0.277278175577521, 0.356421533739194, 0.610177963040769, 0.472000071080402, 0.66534478822723, 0.208317073527724, 0.00724949478171766, 0.144305693684146), actual = c(1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"