Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.8103056        0.7889317 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5237422        0.5237422        0.5217701 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.0814715637825429, 0.985134685412049, 0.739019268658012, 0.295013272203505, 0.614827860379592, 0.60008855490014, 0.510796548333019, 0.436700995778665, 0.275845451746136, 0.560743701644242, 0.963054426712915, 0.149611413944513, 0.685540430713445, 0.944424045039341, 0.200261979363859, 0.753640825860202, 0.473560260608792, 0.0784136403817683, 0.195060894591734, 0.360089509980753, 0.9924593239557, 0.872719965875149, 0.78990541677922, 0.853717490332201, 0.409555808408186, 0.556049634702504, 0.760890644043684, 0.874511415604502, 0.557151281740516, 0.195936338976026, 0.47368831676431, 0.29110126150772, 0.257272963644937, 0.523291698656976, 0.70812057889998, 0.6459514235612, 0.251748387468979, 0.235090498346835, 0.959397217724472, 0.0591722200624645, 0.9222521209158, 0.117284061620012, 0.676445389632136, 0.755755348596722, 0.771591728786007, 0.525504528777674, 0.417354780482128, 0.21566265868023, 0.322142089949921, 0.0910574903246015, 0.764282503863797, 0.840134046738967, 0.08184416545555, 0.394746849546209, 0.353192541515455, 0.592702863970771, 0.901421598391607, 0.291814141906798, 0.782941369805485, 0.400440870085731, 0.245187080930918, 0.564602288184687, 0.64930786495097, 0.275937397964299, 0.376052264124155, 0.0485156700015068, 0.32796791382134, 0.751372473780066, 0.242544115521014, 0.649652891093865, 0.903560754843056, 0.543881558813155, 0.210699345218018, 0.2139748390764, 0.288061454659328, 0.659202445764095, 0.0545273751486093, 0.800983673892915, 0.885035296203569, 0.979646330699325, 0.429346937453374, 0.180677985306829, 0.131528762867674, 0.364015949657187, 0.980090320343152, 0.774086917284876, 0.149503412889317, 0.0115820709615946, 0.36785554792732, 0.0246028832625598, 0.324892353033647, 0.403326894622296, 0.373445329954848, 0.381899152416736, 0.584259754046798, 0.782209926052019, 0.18025859980844, 0.230546009261161, 0.579134640051052, 0.151998597895727, 0.635353011544794, 0.250454191816971, 0.947020109510049, 0.348823821172118, 0.766801233170554, 0.911339909769595, 0.542607386596501, 0.850662148790434, 0.163979397621006, 0.914064307929948, 0.219763378379866, 0.288307606941089, 0.14589116955176, 0.467619417002425, 0.341450165957212, 0.929260611301288, 0.772904676385224, 0.427297920454293, 0.211815628455952, 0.474421395920217, 0.967308065621182, 0.654790223110467, 0.264182348968461, 0.374255338450894, 0.435282152611762, 0.599713960895315, 0.732503122184426, 0.988538048462942, 0.383846117649227, 0.388936523115262, 0.859937550500035, 0.486446822062135, 0.454102350631729, 0.387257790425792, 0.0746732533443719, 0.111635052133352, 0.765153761720285, 0.224219660507515, 0.473829926224425, 0.879560984671116, 0.880963245872408, 0.89930922165513, 0.0609494701493531, 0.997299100505188, 0.453008195385337, 0.781891803257167, 0.467246667249128, 0.381865240400657, 0.0189564030151814, 0.0685273841954768, 0.19872724195011, 0.17589819827117, 0.20994472480379, 0.779323037713766, 0.715191055089235, 0.341313631506637, 0.467883423902094, 0.975129678612575, 0.384216328850016, 0.778743023052812, 0.266351086087525, 0.432564494200051, 0.131949724629521, 0.431178374448791, 0.571123851696029, 0.780399717856199, 0.0346086446661502, 0.296661146217957, 0.572517884429544, 0.668844440253451, 0.233773321378976, 0.612755748676136, 0.499634806299582, 0.125441033393145, 0.387681748950854, 0.231142976786941, 0.357512512709945, 0.1372523275204, 0.339194061467424, 0.0659698566887528, 0.6934779798612, 0.849265459226444, 0.506862396840006, 0.33206352731213, 0.826751337153837, 0.452437849482521, 0.706693272804841, 0.336963503621519, 0.827966715209186, 0.213663772679865, 0.970483201090246, 0.231057989876717, 0.0400124897714704, 0.680421368451789, 0.413535626372322, 0.395199027610943, 0.712865176377818, 0.646400277735665, 0.61099394899793, 0.485817365115508, 0.977512731216848, 0.0353214081842452, 0.314358238596469, 0.503513062139973, 0.0442690849304199, 0.226149827241898, 0.651448469376191, 0.883635765872896, 0.441103767603636, 0.0127091438043863, 0.00350571982562542, 0.21933153970167, 0.446535408496857, 0.92725554225035, 0.017827661242336, 0.948892381275073, 0.111143575282767, 0.269213630119339, 0.446889721089974, 0.239764735335484, 0.414226169930771, 0.465167409041896, 0.878839635290205, 0.204750265460461, 0.22834401531145, 0.661484722280875, 0.0929564177058637, 0.00550819584168494, 0.145934762666002, 0.681128157768399, 0.650340188061818, 0.381214160239324, 0.175876137334853, 0.637192105408758, 0.576420110883191, 0.995202609570697, 0.648010321194306, 0.111923517193645, 0.148525295313448, 0.945694192312658, 0.702076314250007, 0.235510861268267, 0.650687146000564, 0.766764825209975, 0.570247118128464, 0.503938231617212, 0.577855078503489, 0.22516902233474, 0.533815690083429, 0.015314046991989, 0.663878240855411, 0.0962579639162868, 0.842891820240766, 0.154833302134648, 0.334918503882363, 0.321547030936927, 0.947356903227046, 0.363627502229065, 0.795618015574291, 0.719823740655556, 0.450467163696885, 0.927533053560182, 0.967486312845722, 0.925231094937772, 0.183111799880862, 0.413619883591309, 0.379447268787771, 0.156934038503096, 0.307724897284061, 0.428979845019057, 0.578655199147761, 0.764724271604791, 0.626895413501188, 0.942051796009764, 0.248642277438194, 0.209918612847105, 0.229477487038821, 0.913812408223748, 0.937039614189416, 0.0219888682477176, 0.124930708669126, 0.501227562315762, 0.0814190113451332, 0.423714857082814, 0.464649168541655, 0.410344733856618, 0.0877576023340225, 0.0599900521337986, 0.707194344606251, 0.281322030117735, 0.526779879583046, 0.50778863299638, 0.787799273384735, 0.8321972056292, 0.536550760734826, 0.137697234051302, 0.884018780430779, 0.828447556821629, 0.431047562975436, 0.848413143772632, 0.890733996871859, 0.818548974813893, 0.666278813267127, 0.511448420118541, 0.427916299551725, 0.528687794459984, 0.991183320526034, 0.65734004881233, 0.841698939912021, 0.6559166777879, 0.859661003574729, 0.707964684814215, 0.157406129641458, 0.701788823818788, 0.218710767803714, 0.32934658951126, 0.934526932192966, 0.781561040086672, 0.431706708623096, 0.744402726646513, 0.105533228022978, 0.993797381874174, 0.00805320637300611, 0.566591790178791, 0.887057965388522, 0.66038340004161, 0.686680364189669, 0.0893969375174493, 0.982113608391955, 0.417819106020033, 0.385417859535664, 0.63266045646742, 0.543379276525229, 0.776232266565785, 0.323414185550064, 0.433177801081911, 0.929438307648525, 0.890921907266602, 0.833578376332298, 0.238433384103701, 0.403416835935786, 0.667744912439957, 0.968855397542939, 0.769719010451809, 0.424663071520627, 0.0339336327742785, 0.543992919148877, 0.0820885626599193, 0.887016491265967, 0.337409669999033, 0.234838014934212, 0.0536450655199587, 0.34816297586076, 0.732501113787293, 0.0462994806002825, 0.721444626571611, 0.761171532329172, 0.469385240226984, 0.849657788174227, 0.726924969349056, 0.959171225316823, 0.65112894657068, 0.188257334055379, 0.719665811397135, 0.582874022657052, 0.872490161564201, 0.294714155606925, 0.211597410961986, 0.11002642614767, 0.877460540970787, 0.735689944354817, 0.902679006336257, 0.120957831619307, 0.0655564705375582, 0.547791090328246, 0.746968064457178, 0.219047360820696, 0.416394415311515, 0.357511654263362, 0.27397980960086, 0.124110729433596, 0.682798760477453, 0.46733284345828, 0.817213360453025, 0.892955130897462, 0.0852088469546288, 0.266407467192039, 0.893141106236726, 0.996585120679811, 0.874344725860283, 0.817574326181784, 0.908835670445114, 0.84766928688623, 0.489031085278839, 0.898507150355726, 0.828779714880511, 0.463182060746476, 0.213520449586213, 0.886234532110393, 0.836432387586683, 0.877544510411099, 0.737758046016097, 0.825183116365224, 0.587887042202055, 0.399439189350232, 0.0828753078822047, 0.289447188843042, 0.338106602430344, 0.387432996183634, 0.709274496883154, 0.154563193907961, 0.758027039235458, 0.346249523339793, 0.388570396928117, 0.759290844202042, 0.761941473931074, 0.97335758456029, 0.285295962356031, 0.180641202023253, 0.69754693005234, 0.674618889577687, 0.820094655733556, 0.857067379401997, 0.023982274113223, 0.0955227005761117, 0.888934635091573, 0.248400507029146, 0.622521981829777, 0.704882911406457, 0.744065060745925, 0.472229480510578, 0.199380608275533, 0.280351636465639, 0.525989510584623, 0.420731569407508, 0.018261646386236, 0.752721201628447, 0.558231877163053, 0.203153690556064, 0.776082299416885, 0.534675938077271, 0.804699299391359, 0.0253708029631525, 0.689590863650665, 0.321676089195535, 0.94229571102187, 0.651328723877668, 0.385718951933086, 0.415771214757115, 0.91698454762809, 0.342040555551648, 0.322286161594093, 0.720796292647719, 0.0131081559229642, 0.433807814260945, 0.684856145409867, 0.801992834778503, 0.719405028037727, 0.218165755970404, 0.452528701396659, 0.0598739823326468, 0.0703333436977118, 0.80012292182073, 0.29574795300141, 0.346446088515222, 0.484582061646506, 0.98087759828195, 0.895219459198415, 0.0972888162359595, 0.880374152446166, 0.0288411735091358, 0.426023448351771, 0.507818768499419, 0.506571989273652, 0.775107970461249, 0.600642723264173, 0.148891282035038, 0.532955821603537, 0.513336258009076, 0.966887810500339, 0.60898221982643, 0.105651227291673, 0.687941359123215, 0.684633734868839, 0.662422031862661, 0.539395779836923, 0.267745797988027, 0.184271132107824, 0.831652168184519, 0.869593715760857, 0.0866942701395601, 0.793859715107828, 0.437083488330245, 0.42015089886263, 0.896759560098872, 0.585612545721233, 0.392808982403949, 0.248477563494816, 0.924330091103911, 0.715816357173026, 0.350521618267521, 0.220523838652298, 0.822437313618138, 0.286153642227873, 0.902479330077767, 0.432184956502169, 0.212435666238889, 0.708899380872026, 0.107183643849567, 0.0372502999380231, 0.432768175378442, 0.819886962650344, 0.688743017613888, 0.355328198289499, 0.0309125285129994, 0.103431810624897, 0.795166235649958, 0.717701592249796, 0.872490285430104, 0.542973417090252, 0.0285625525284559, 0.641879095695913, 0.889608269324526, 0.666569580789655, 0.964687153464183, 0.305228021228686, 0.906553355278447, 0.193989590508863, 0.62814134801738, 0.262227698462084, 0.41144264372997, 0.184066673507914, 0.200638329843059, 0.388711643638089, 0.959669516421854, 0.120731297647581, 0.710796999977902, 0.0658151444513351, 0.141767865046859, 0.870713122887537, 0.716648146510124, 0.305576970567927, 0.880349568091333, 0.125899226171896, 0.698873971123248, 0.503470263443887, 0.447954430244863, 0.20488191023469, 0.583578233839944, 0.991411745082587, 0.899186300346628, 0.893374972511083, 0.527133083436638, 0.369322885060683, 0.722817232133821, 0.625998475123197, 0.359538102522492, 0.404071918921545, 0.265996247762814, 0.301255922531709, 0.0114454748108983, 0.0125314677134156, 0.940730486530811, 0.213966104434803, 0.502997531788424, 0.976518294308335, 0.946801146492362, 0.887088859919459, 0.672581739956513, 0.451485156081617, 0.252357123885304, 0.382011448265985, 0.345985814463347, 0.216934810159728, 0.631293470039964, 0.705617622006685, 0.625538953579962, 0.376107508083805, 0.117308742366731, 0.407684026751667, 0.402582926442847, 0.181110532488674, 0.854199854424223, 0.931072463747114, 0.10861867852509, 0.151420163922012, 0.702511909650639, 0.752285014139488, 0.383241950301453, 0.0507367986720055, 0.421020644716918, 0.382229494862258, 0.441805227426812, 0.246531958691776, 0.946865733945742, 0.20080257137306, 0.35438588866964, 0.213363313348964, 0.712449064711109, 0.571344724623486, 0.208687086822465, 0.357962256995961, 0.736656729830429, 0.170819797553122, 0.106884684180841, 0.656535341171548, 0.982678452273831, 0.0222918344661593, 0.0031430299859494, 0.17715810239315, 0.038382368395105, 0.50115429982543, 0.058272534282878, 0.21403779508546, 0.844830173300579, 0.669931738404557, 0.934438064461574, 0.532135291257873, 0.276530202012509, 0.618844471639022, 0.0126420157030225, 0.639671710785478, 0.408648864366114, 0.652299862587824, 0.0957135586068034, 0.045161898015067, 0.630463647656143, 0.811058104271069, 0.992397021735087, 0.111388524528593, 0.647257063537836, 0.695802351692691, 0.98297670437023, 0.0305808747652918, 0.64910470880568, 0.0926812966354191, 0.543027964653447, 0.81490686791949, 0.651341846212745, 0.751229151384905, 0.73622944811359, 0.38902216590941, 0.448925700737163, 0.866905224276707, 0.927374853054062, 0.655263302614912, 0.640811735065654, 0.201038544531912, 0.0393482293002307, 0.411837431835011, 0.964795677689835, 0.417274414561689, 0.725821082480252, 0.297153632156551, 0.489750117296353, 0.739786230027676, 0.592406455427408, 0.468531557824463, 0.968835230451077, 0.836137340171263, 0.129364460706711, 0.238125869305804, 0.749859847594053, 0.0392297992948443, 0.366651566699147, 0.693993324879557, 0.327103741466999, 0.840618563117459, 0.69479588768445, 0.961132695898414, 0.48185673635453, 0.0347718251869082, 0.045369815081358, 0.722282590577379, 0.440851433202624, 0.999817050294951, 0.63045864389278, 0.352050596382469, 0.510070101125166, 0.514822808094323, 0.768050760263577, 0.441476396284997, 0.355477260891348, 0.366236818954349, 0.974579143570736, 0.588582962518558, 0.63909661071375, 0.0382750672288239, 0.108896268066019, 0.93779333611019, 0.0944837806746364, 0.0157937970943749, 0.569522001314908, 0.987454069778323, 0.0855581504292786, 0.826390990754589, 0.133607825497165, 0.313008837634698, 0.769762403098866, 0.538796821609139, 0.421670086216182, 0.926416706526652, 0.161600963911042, 0.409608444431797, 0.178364175138995, 0.728593123145401, 0.253271655412391, 0.991033807164058, 0.93561834958382, 0.700120940804482, 0.805531826335937, 0.369090423453599, 0.770110575016588, 0.877843894297257, 0.655900059267879, 0.60443653492257, 0.164391162106767, 0.5113545502536, 0.877536623505875, 0.273512234445661, 0.58979009417817, 0.501240646000952, 0.584611715283245, 0.607647263212129, 0.299435259075835, 0.401213015895337, 0.426971902139485, 0.740743614034727, 0.799281865824014, 0.893036694964394, 0.266476993681863, 0.579617600422353, 0.552834733622149, 0.910777243087068, 0.0820949859917164, 0.00344584858976305, 0.0691236625425518, 0.852010363712907, 0.343549170531332, 0.544074046891183, 0.524271039990708, 0.0160307807382196, 0.300727906636894, 0.193748139543459, 0.890767617616802, 0.217742971377447, 0.437338207615539, 0.475907088723034, 0.683891089633107, 0.154637838015333, 0.929722619010136, 0.421937962528318, 0.689879144076258, 0.977110935607925, 0.563487611012533, 0.562468799529597, 0.0359893925487995, 0.0140506580937654, 0.668089133221656, 0.0913872469682246, 0.71498463395983, 0.0815451715607196, 0.355997786158696, 0.765224273316562, 0.459883835632354, 0.298699144739658, 0.341875144280493, 0.105154064483941, 0.819764762418345, 0.551368853310123, 0.105782819446176, 0.350321055389941, 0.418170303571969, 0.475178331602365, 0.77082234271802, 0.72122057643719, 0.969263914274052, 0.762161808088422, 0.625144728226587, 0.44431432033889, 0.226934185018763, 0.728658980689943, 0.666480348212644, 0.675052866572514, 0.378004644997418, 0.78873254545033, 0.748204086674377, 0.358576306840405, 0.735761053394526, 0.790409250184894, 0.520167669281363, 0.589264240814373, 0.313048144802451, 0.0729163773357868, 0.0965472508687526, 0.623905327403918, 0.327965590870008, 0.849230115069076, 0.651657934300601, 0.354226208059117, 0.384673655731604, 0.309666501590982, 0.605243237456307, 0.421509504551068, 0.75270326109603, 0.500589940696955, 0.371700452407822, 0.528212541248649, 0.550013903994113, 0.141887816367671, 0.73961920524016, 0.803333391202614, 0.0649077533744276, 0.345827665412799, 0.767215495929122, 0.594677413348109, 0.924887154484168, 0.0469145043753088, 0.765169329475611, 0.0393976855557412, 0.392116298433393, 0.81280814553611, 0.368244109675288, 0.819455121876672, 0.190469406777993, 0.603102437686175, 0.671752190450206, 0.44105703337118, 0.313706032000482, 0.901003305567428, 0.648953229887411, 0.107218928867951, 0.733121334109455, 0.745269876206294, 0.282271689269692, 0.680869841948152, 0.525158162461594, 0.433489907998592, 0.00160313118249178, 0.179089265875518, 0.718679701443762, 0.850968768354505, 0.650446712505072, 0.755319810239598, 0.517698733136058, 0.673115604789928, 0.124861539108679, 0.659960362827405, 0.632462195586413, 0.659301718696952, 0.854439920978621, 0.0380921184550971, 0.0879822648130357, 0.408179916907102, 0.578855886589736, 0.300770781235769, 0.928011654410511, 0.753439658554271, 0.274899545125663, 0.977910767542198, 0.694358224514872, 0.786349969450384, 0.720288962358609, 0.168229593429714, 0.214904542546719, 0.0563892927020788, 0.543122032424435, 0.559527657227591, 0.963409054325894, 0.0230963050853461, 0.927056372398511, 0.705990836024284, 0.284316636156291, 0.0702850809320807, 0.969947643345222, 0.166790721472353, 0.93819740624167, 0.773373294156045, 0.109826517058536, 0.824511768296361, 0.624981458997354, 0.878507357789204, 0.844883522717282, 0.43042221176438, 0.666710804682225, 0.150747735751793, 0.775163231650367, 0.555763041134924, 0.0455414424650371, 0.759847238427028, 0.225483912974596, 0.977864264277741, 0.710243315668777, 0.678110407199711, 0.614277406595647, 0.438609804259613, 0.427782843820751, 0.523201808799058, 0.694802204146981, 0.661722911056131, 0.616888858843595, 0.538982615806162, 0.680964722996578, 0.962649396853521, 0.526130140759051, 0.0687208161689341, 0.298689099494368, 0.267830695491284, 0.380254059331492, 0.185613648267463, 0.955889286007732, 0.737545932177454, 0.782590222312137, 0.562296376330778, 0.440567278768867, 0.10787648614496, 0.180406887084246, 0.166495860554278, 0.0761723010800779, 0.935036676935852, 0.235008330550045, 0.708605004707351, 0.676904706517234, 0.62222500750795, 0.728522032964975, 0.534256456885487, 0.728596830973402, 0.420947863487527, 0.610160305164754, 0.0769826821051538, 0.221941211493686, 0.0113223064690828, 0.479464509291574, 0.745720993261784, 0.733891352079809, 0.657216419233009, 0.0407988561782986, 0.376145764719695, 0.118026313604787, 0.187895509880036, 0.116843120660633, 0.345254501793534, 0.902697418816388, 0.715853159315884, 0.512941062217578, 0.157134319655597, 0.609328293474391, 0.383131434209645, 0.909666650928557, 0.913213715422899, 0.53918032348156, 0.352920430712402, 0.438974467804655, 0.323156009195372, 0.857652871171013, 0.578299593646079, 0.985229002311826, 0.192641470348462, 0.00257103447802365, 0.360468605067581, 0.761279022786766, 0.891374872066081, 0.370521703734994, 0.0475128244142979, 0.976518517127261, 0.937507985858247, 0.875241010216996, 0.649631556123495, 0.413315910380334, 0.89980850322172, 0.281347404001281, 0.330402155406773, 0.475837260251865, 0.491836224449798, 0.768060240428895, 0.0791088510304689, 0.952781184343621, 0.477909575914964, 0.0389334631618112, 0.805700783850625, 0.235984557773918, 0.953462132718414, 0.0223193948622793, 0.752409436041489, 0.185783953173086, 0.974710995098576, 0.634992242557928, 0.52548817358911, 0.0994589778129011, 0.305540229193866, 0.463193953270093, 0.69144614553079, 0.751655236585066, 0.890393851092085, 0.340099944965914, 0.960896236589178, 0.119495312916115, 0.851891340687871, 0.466541982023045, 0.468590859090909), actual = c(1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.91350429598242, 0.445288447663188, 0.404796838294715, 0.0251869233325124, 0.170629874803126, 0.973676682682708, 0.0607925422955304, 0.756157144438475, 0.684077570447698, 0.765242757974193, 0.525486525148153, 0.478194320574403, 0.267508211312816, 0.239647550974041, 0.330128558678553, 0.68633020715788, 0.184507004916668, 0.514258336974308, 0.345451112603769, 0.535395285114646, 0.664220592705533, 0.173738296842203, 0.957773570902646, 0.0923262154683471, 0.696913028135896, 0.0274976091459394, 0.0298727799672633, 0.0156643786467612, 0.295008231187239, 0.145759992999956, 0.24149907194078, 0.0446367391850799, 0.228173745563254, 0.729145858436823, 0.827047908678651, 0.758155361749232, 0.923914730548859, 0.460064722457901, 0.847028113203123, 0.531758920988068, 0.23305632523261, 0.556972078280523, 0.0713968621566892, 0.301707074744627, 0.608779686037451, 0.432208913611248, 0.291750009637326, 0.219682646915317, 0.564349783118814, 0.64855127572082, 0.733251454308629, 0.237184163182974, 0.901680963812396, 0.528176572639495, 0.72443431452848, 0.787645169533789, 0.430908338632435, 0.459808025276288, 0.967176462523639, 0.811188956024125, 0.122138110222295, 0.543266030726954, 0.41133320145309, 0.318717147922143, 0.557048430200666, 0.167032066034153, 0.381783706601709, 0.330607350217178, 0.959588012192398, 0.962858560262248, 0.675977140432224, 0.856140187010169, 0.109824103303254, 0.735411582980305, 0.985965220024809, 0.122987763257697, 0.732655743602663, 0.645409919787198, 0.178665474290028, 0.177593914559111, 0.369386154226959, 0.280242208158597, 0.181400386616588, 0.992215994745493, 0.38194827735424, 0.72879465483129, 0.0696661209221929, 0.0690818440634757, 0.0487203623633832, 0.585994202643633, 0.43122993898578, 0.945632570190355, 0.544036214705557, 0.443355498602614, 0.997679801424965, 0.90670566726476, 0.114063906483352, 0.767219881759956, 0.0349399703554809, 0.532827551476657, 0.436922087101266, 0.789753741119057, 0.966901862062514, 0.904242525575683, 0.906871975399554, 0.352705427212641, 0.0298244752921164, 0.156893038656563, 0.773484160657972, 0.915051374817267, 0.947556793224066, 0.738102749688551, 0.846486589172855, 0.300032007042319, 0.502589830663055, 0.285647382028401, 0.0106456982903183, 0.868814778979868, 0.6585986581631, 0.243795190472156, 0.350547060836107, 0.470520371338353, 0.168846849119291, 0.422716414788738, 0.867764089722186, 0.310503570595756, 0.171159421792254, 0.493257481139153, 0.89359721285291, 0.696179699152708, 0.174465109361336, 0.0297219571657479, 0.962113078217953, 0.269769568461925, 0.517205490265042, 0.94677363242954, 0.360356139717624, 0.567392746917903, 0.286112726898864, 0.914413141552359, 0.363858515163884, 0.919824557146057, 0.855629554949701, 0.0913653490133584, 0.980270521715283, 0.711231819121167, 0.220188393257558, 0.693261067848653, 0.927942908369005, 0.454795113997534, 0.871212074998766, 0.55889022978954, 0.632049719570205, 0.182115028379485, 0.952667549485341, 0.587399620795622, 0.323989769909531, 0.128872341942042, 0.206588551634923, 0.34965897561051, 0.0957829772960395, 0.37952498649247, 0.303176561836153, 0.356393988477066, 0.125867311377078, 0.0802027778699994, 0.124724110588431, 0.785712039563805, 0.875783620402217, 0.330110559472814, 0.682795844506472, 0.500237054657191, 0.0708311614580452, 0.341149680316448, 0.912017110735178, 0.999981683911756, 0.436809421051294, 0.415423459140584, 0.0243870816193521, 0.429711413569748, 0.178615210577846, 0.34609440411441, 0.561088925693184, 0.386513700243086, 0.915572237456217, 0.909878653939813, 0.787150621879846, 0.644498590379953, 0.353231705026701, 0.317844139644876, 0.892023088177666, 0.555141485994682, 0.59574609878473, 0.268229647073895, 0.281452844385058, 0.597040058812127, 0.700681714806706, 0.706529427552596, 0.980144861154258, 0.685676334891468, 0.902944878675044, 0.340827066916972, 0.984860515687615, 0.686790909385309, 0.20632182364352, 0.555832711514086, 0.0647059923503548, 0.716143170837313, 0.115103788906708, 0.102439352776855, 0.574383236700669, 0.217434195103124, 0.115728434408084, 0.0695200846530497, 0.854381793178618, 0.695513624465093, 0.186779049923643, 0.625324634369463, 0.237046058988199, 0.385258501162753, 0.208840230247006, 0.650094133568928, 0.805186672601849, 0.602944107726216, 0.819134945049882, 0.817086707567796, 0.590697393054143, 0.788199543021619, 0.59243930503726, 0.410558337811381, 0.454190128250048, 0.979218075051904, 0.57530958764255, 0.330648412695155, 0.18209506990388, 0.138668314088136, 0.4395725002978, 0.247099150205031, 0.591406987048686, 0.791258124401793, 0.234416452934965, 0.727566034300253, 0.206055159447715, 0.810549716232345, 0.549694555578753, 0.0770418476313353, 0.155365490121767, 0.0801388737745583, 0.696819243486971, 0.397527455817908, 0.426687917672098, 0.892833903199062, 0.367173149716109, 0.852150975726545, 0.904123600572348, 0.0528911594301462, 0.0343045580666512, 0.774405439849943, 0.801756754983217, 0.157432518200949, 0.0294948157388717, 0.444325432414189, 0.702416467480361, 0.374181446619332, 0.134493401739746, 0.0737586976028979, 0.705007223412395, 0.0825219266116619, 0.799247332848608, 0.821972555015236, 0.0638382367324084, 0.902570771053433, 0.800535873742774, 0.73486697091721, 0.336153647862375, 0.485445365309715, 0.161532415542752, 0.130553272319958, 0.924606463639066, 0.614340793574229, 0.150465056998655, 0.863856148673221, 0.648947683628649, 0.653392548905686, 0.220608888659626, 0.780727211153135, 0.656184817198664, 0.871783891227096, 0.320125826401636, 0.397210255265236, 0.267906895373017, 0.921494874171913, 0.210704541066661, 0.519149176077917, 0.198214155156165, 0.10621554707177, 0.592071480350569, 0.252125795232132, 0.396273895632476, 0.621123351156712, 0.47014017496258, 0.297278065234423, 0.585974137065932, 0.850054983980954, 0.244719913462177, 0.306216312106699, 0.196750008035451, 0.206227479735389, 0.51541706523858, 0.503054061206058, 0.724687597248703, 0.904977944446728, 0.749462192179635, 0.514672162244096, 0.226028806529939, 0.972251021768898, 0.946478729369119, 0.371056923875585, 0.86413516686298, 0.39718989119865, 0.304091286612675, 0.896797408815473, 0.589918657904491, 0.914066356373951, 0.448376068845391, 0.69282273738645, 0.982441013678908, 0.0640385516453534, 0.0813003717921674, 0.288856533588842, 0.834099577507004, 0.442771235713735, 0.159485289128497, 0.490896773990244, 0.977655343012884, 0.724517793860286, 0.870620410656556, 0.470516374800354, 0.0802673250436783, 0.226595426676795, 0.324986105551943, 0.895213864510879, 0.918995559215546, 0.33448501303792, 0.339102409314364, 0.0198003763798624, 0.512559955939651, 0.134216827107593, 0.681754909222946, 0.180793792940676, 0.651085518766195, 0.513181953690946, 0.313357242615893, 0.92826225142926, 0.127212062710896, 0.927200630074367, 0.756315907696262, 0.337010088376701, 0.932068517897278, 0.146405629813671, 0.539032425032929, 0.541923146694899, 0.000714094843715429, 0.761967671103776, 0.404920817585662, 0.286903015570715, 0.0805590280797333, 0.489508215570822, 0.123624861938879, 0.692621986847371, 0.88528325362131, 0.300353951985016, 0.767356431111693, 0.11181920976378, 0.315646771807224, 0.752087198663503, 0.883256172761321, 0.796085519017652, 0.39491676678881, 0.184623023029417, 0.585723492782563, 0.832523768534884, 0.527219721814618, 0.154145488515496, 0.280492710415274, 0.102863904088736, 0.720282072667032, 0.378511732444167, 0.93862702534534, 0.637864719843492, 0.47858249861747, 0.549752495251596, 0.617677521426231, 0.001039057970047, 0.630802099592984, 0.987931350711733, 0.222707096254453, 0.425546505255625, 0.207450055982918, 0.195736096706241, 0.564897385658696, 0.389625374227762, 0.835437068715692, 0.333939174888656, 0.366123937536031, 0.850214435020462, 0.303590500494465, 0.316249638097361, 0.962559985928237, 0.141279496019706, 0.52976606390439, 0.277063561370596, 0.168740993831307, 0.208350214874372, 0.663945802487433, 0.389583266107365, 0.0641673766076565, 0.225617381511256, 0.70368801523, 0.594125512987375, 0.743778771720827, 0.863445872208104, 0.160085965879261, 0.12843711161986, 0.305994888069108, 0.313886357704177, 0.348461571149528, 0.214551977347583, 0.305770576931536, 0.576760560506955, 0.215898313326761, 0.225285874912515, 0.742971424479038, 0.972712985007092, 0.186588933691382, 0.499476059805602, 0.538163991179317, 0.677555952686816, 0.716849635355175, 0.136095002759248, 0.100765213137493, 0.682968629524112, 0.649457732681185, 0.81622620834969, 0.889162319712341, 0.942513993708417, 0.757151958532631, 0.953778359806165, 0.791997727239504, 0.56102193868719, 0.279221750563011, 0.8352992371656, 0.895658491877839, 0.252423835685477, 0.290303847985342, 0.342337408103049, 0.276257694233209, 0.910909577272832, 0.0642072567716241, 0.00909616844728589, 0.598830412840471, 0.878180428873748, 0.563886162126437, 0.804540546843782, 0.842876971932128, 0.462139870971441, 0.577881795121357, 0.160265740239993, 0.830191398970783, 0.791789701441303, 0.46218606014736, 0.774051481392235, 0.434625662630424, 0.719813120784238, 0.721983114723116, 0.518861796474084, 0.800415447214618, 0.15843594330363, 0.135449356865138, 0.384776699123904, 0.846531891031191, 0.528803861699998, 0.003384716110304, 0.113752242643386, 0.898847398348153, 0.507880639983341, 0.241385566536337, 0.598232734017074, 0.560054045636207, 0.711940319743007, 0.908378683496267, 0.792445437051356, 0.918943743687123, 0.937660408206284, 0.370773564325646, 0.978291475679725, 0.130212849238887, 0.620432394789532, 0.47471277997829, 0.0507675709668547, 0.636805576272309, 0.0644131517037749, 0.262876101769507, 0.82474353373982, 0.481225552037358, 0.973085438134149, 0.911856636404991, 0.371559824096039, 0.962117731804028, 0.130570439854637, 0.537656541913748, 0.192630771780387, 0.969616821501404, 0.824746168917045, 0.69225610489957, 0.974869070807472, 0.60049238242209, 0.0764143862761557, 0.275038512190804, 0.0721240425482392, 0.54016328509897, 0.435852165566757, 0.126617534086108, 0.834347016410902, 0.0170266937930137, 0.0564201767556369, 0.687361924443394, 0.279639245243743, 0.508218475617468, 0.577678193803877, 0.95579758589156, 0.655784333124757, 0.551536780549213, 0.86433965805918, 0.661700916476548, 0.841142894001678, 0.441063263453543, 0.786539874970913, 0.829098992049694, 0.128365229116753, 0.829793806187809, 0.139270306564867, 0.291577008087188, 0.403368857689202, 0.0422487007454038, 0.162432278972119, 0.985666459891945, 0.592238091863692, 0.148968996945769, 0.258895973209292, 0.958544079214334, 0.76864296826534, 0.868673973716795, 0.658111942466348, 0.286136141978204, 0.790645056869835, 0.432803303934634, 0.115508298389614, 0.213649999117479, 0.515845258720219, 0.882830460788682, 0.644456540001556, 0.532638255041093, 0.149091411149129, 0.517575775738806, 0.0712897803168744, 0.2186986124143, 0.899271475849673, 0.760999171761796, 0.937563120154664, 0.1099731719587, 0.880167413270101, 0.69120795535855, 0.193385078338906, 0.0358024993911386, 0.752695981413126, 0.306507652625442, 0.800221518380567, 0.856419842224568, 0.0734828850254416, 0.697368650930002, 0.445751396706328, 0.286687299376354, 0.771189289167523, 0.791133670601994, 0.910055447835475, 0.989656188990921, 0.780784906586632, 0.200725161004812, 0.836151237366721, 0.536667846841738, 0.790432530455291, 0.558906942605972, 0.293725385330617, 0.278639637865126, 0.211354354862124, 0.138888941844925, 0.246409619925544, 0.43979943334125, 0.158808259060606, 0.364236725494266, 0.17889763484709, 0.609102842165157, 0.0237957229837775, 0.563520007533953, 0.859910567989573, 0.26739181834273, 0.220340816536918, 0.988488969625905, 0.676893699681386, 0.00421727378852665, 0.130818268284202, 0.503305728547275, 0.283528911881149, 0.0887105637229979, 0.106800271896645, 0.540589641546831, 0.898897054605186, 0.241407784167677, 0.96735946694389, 0.939128401456401, 0.054616357665509, 0.961122163338587, 0.605513648595661, 0.350807737326249, 0.660826245322824, 0.823041027411819, 0.446026744088158, 0.406981518259272, 0.0938905554357916, 0.495963777648285, 0.320048898458481, 0.771739798830822, 0.48616552259773, 0.142967184539884, 0.380115272011608, 0.326648862101138, 0.117635044734925, 0.452627248130739, 0.669408594723791, 0.857102896319702, 0.539051806088537, 0.885779978707433, 0.511174108833075, 0.725657312897965, 0.35945243970491, 0.989815880777314, 0.525601064087823, 0.448930299840868, 0.59470999869518, 0.928459489019588, 0.106182059738785, 0.155117800692096, 0.630475442390889, 0.161918122088537, 0.433302190387622, 0.194065160583705, 0.734175729798153, 0.93167885649018, 0.718236234039068, 0.952378751011565, 0.723380488809198, 0.912201035534963, 0.474810648476705, 0.748154571978375, 0.676541850436479, 0.489535712869838, 0.729477075161412, 0.563159539364278, 0.0209196687210351, 0.574708193540573, 0.224953999044374, 0.12505357968621, 0.350444010924548, 0.0608523353002965, 0.134317096089944, 0.698383000912145, 0.184656235156581, 0.66826104465872, 0.521490747341886, 0.404595303349197, 0.0853159411344677, 0.295214993646368, 0.578917857725173, 0.337824585149065, 0.0454588462598622, 0.212152414256707, 0.411015228135511, 0.424371627392247, 0.13716505584307, 0.182430249406025, 0.611312596825883, 0.523081606719643, 0.261091063963249, 0.69279291620478, 0.0405344080645591, 0.419235874898732, 0.433246451430023, 0.882480761734769, 0.511937463656068, 0.987340592313558, 0.0853511220775545, 0.914192558266222, 0.507934115594253, 0.565646388102323, 0.828616829123348, 0.285582701675594, 0.279267117381096, 0.020081264199689, 0.907764870440587, 0.31454354012385, 0.289244367275387, 0.183692235965282, 0.843549572397023, 0.356793188257143, 0.543760166270658, 0.0251225142274052, 0.612540177069604, 0.232535346178338, 0.765688456594944, 0.394228271907195, 0.783085400471464, 0.0992450609337538, 0.711239363532513, 0.169489614898339, 0.977681897813454, 0.984278437448665, 0.585774736013263, 0.981868402101099, 0.0409328213427216, 0.988948141224682, 0.528941719792783, 0.523748701205477, 0.173813544213772, 0.48313115327619, 0.310973945539445, 0.660301213385537, 0.715878309449181, 0.179675874300301, 0.525244208285585, 0.24827265040949, 0.718291470082477, 0.738924568984658, 0.86459665116854, 0.74347566999495, 0.288882985711098, 0.59708218392916, 0.969679075293243, 0.73752335482277, 0.913368412991986, 0.6202100624796, 0.189874222269282, 0.496742276242003, 0.28378749312833, 0.130787185626104, 0.0357875279150903, 0.819626126904041, 0.560249934904277, 0.811327122151852, 0.857054653344676, 0.113160205306485, 0.322385706007481, 0.50049909716472, 0.0103418922517449, 0.435390900354832, 0.428744664648548, 0.441908623557538, 0.0789201131556183, 0.585758639033884, 0.179169367998838, 0.226132669486105, 0.751272680936381, 0.731801894493401, 0.340078978100792, 0.267425118014216, 0.0542628685943782, 0.00274266698397696, 0.633733770577237, 0.992726192111149, 0.354703095974401, 0.63551891525276, 0.11796307307668, 0.0301787694916129, 0.620986267924309, 0.466014543548226, 0.304491642862558, 0.636477737221867, 0.479338928824291, 0.341430278029293, 0.859223452396691, 0.438737970078364, 0.306101988302544, 0.676003237487748, 0.814906789688393, 0.795015180716291, 0.809107859386131, 0.807064057793468, 0.136364623205736, 0.268714175792411, 0.655504919355735, 0.181739597581327, 0.636790076736361, 0.651113438652828, 0.220359630184248, 0.220988767920062, 0.572940705809742, 0.377946660621092, 0.707731962902471, 0.628491969313473, 0.153520018793643, 0.740038162097335, 0.0644516381435096, 0.825054952641949, 0.77487807860598, 0.471971468767151, 0.496090272441506, 0.962294190889224, 0.868892756290734, 0.347110596485436, 0.512335455277935, 0.447990009561181, 0.115013092523441, 0.963917586952448, 0.660909721395001, 0.300243219360709, 0.224876824999228, 0.207053993595764, 0.794195271795616, 0.819419412408024, 0.132944622309878, 0.882457702420652, 0.636652423534542, 0.501953833270818, 0.763690445339307, 0.724427549168468, 0.11294287443161, 0.0691930176690221, 0.364019345724955, 0.847001380519941, 0.1499428905081, 0.534574636723846, 0.788279428845271, 0.537017863010988, 0.34380133706145, 0.42984629352577, 0.495009917300195, 0.509381340118125, 0.807333957636729, 0.624296702211723, 0.934188332175836, 0.460205032024533, 0.878506423905492, 0.275639465311542, 0.702577455434948, 0.713159868726507, 0.344574474962428, 0.981271151686087, 0.276074037188664, 0.335180015768856, 0.712303089676425, 0.563971515744925, 0.0455204450991005, 0.69180316850543, 0.0562927727587521, 0.198430730495602, 0.140035908669233, 0.760810380103067, 0.422183588147163, 0.173564186552539, 0.823563710786402, 0.171818433329463, 0.529467051383108, 0.0157766910269856, 0.316971140913665, 0.908745337976143, 0.49266734928824, 0.0392983937636018, 0.569235384231433, 0.414852207526565, 0.632750238757581, 0.716197804547846, 0.977658049203455, 0.6251970322337, 0.955148227047175, 0.122372447280213, 0.911198766436428, 0.984010725747794, 0.692873525898904, 0.582730987807736, 0.67170942784287, 0.412180615589023, 0.939566482556984, 0.734831037232652, 0.572825186653063, 0.737580279586837, 0.748592578107491, 0.508358755148947, 0.208490697899833, 0.76201834436506, 0.133385081775486, 0.489116363925859, 0.101972631644458, 0.247188292210922, 0.91172539582476, 0.688087621470913, 0.546652103308588, 0.494396536145359, 0.852472520666197, 0.514408667804673, 0.995906123192981, 0.338718095095828, 0.216421701246873, 0.861429720651358, 0.194787678075954, 0.713913116138428, 0.799882953753695, 0.98802729207091, 0.648166649742052, 0.710919539211318, 0.744652255671099, 0.177405798109248, 0.119238892802969, 0.475843062857166, 0.817095230100676, 0.522166967159137, 0.155104625970125, 0.351731037953869, 0.934344019507989, 0.523887342074886, 0.341932825278491, 0.892946175532416, 0.424827078357339, 0.182599478168413, 0.830192451830953, 0.764496102463454, 0.687862172955647, 0.912194015923887, 0.936258948408067, 0.889309314079583, 0.0128397648222744, 0.866477646166459, 0.00952909141778946, 0.205189294647425, 0.856304266490042, 0.0580567438155413, 0.202720870496705, 0.165069753304124, 0.208277476020157, 0.485411934088916, 0.259869267232716, 0.160776548553258, 0.0737654047552496, 0.285279867472127, 0.929216095712036, 0.231061416910961, 0.0986290175933391, 0.887604160001501, 0.665019347798079, 0.893924563191831, 0.689891738584265, 0.191803370369598, 0.977413648972288, 0.505779565079138, 0.949119094060734, 0.894300364190713, 0.175704024033621, 0.265745356678963, 0.387973450822756, 0.561580903362483, 0.26215763320215, 0.405987154692411, 0.349198434967548, 0.170885375002399, 0.490430367179215, 0.593212569132447, 0.61914693727158, 0.694008688209578, 0.641365805873647, 0.582873733714223, 0.312698299065232, 0.196082635084167, 0.534055596450344, 0.319406403927132, 0.0878862275276333, 0.775171588407829, 0.963510418543592, 0.501283702906221, 0.46827659313567, 0.795113407541066, 0.186570764984936, 0.226757072145119, 0.595594663405791, 0.521347493631765, 0.189987107180059, 0.759584984742105, 0.736089578829706, 0.408632647246122, 0.951309928437695, 0.170695012900978, 0.652567456709221), actual = c(1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"