Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.3992641        0.3992641 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.3992641        0.3992641        0.4828222 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.211132261902094, 0.689219429157674, 0.862868359312415, 0.54051336622797, 0.201563189039007, 0.591358493315056, 0.779075508238748, 0.184697592863813, 0.881176474504173, 0.200799060752615, 0.799484639428556, 0.4997638175264, 0.676538368919864, 0.88605478964746, 0.934739085612819, 0.930927121313289, 0.723750835517421, 0.134919038973749, 0.440855185734108, 0.0803295772057027, 0.608718682778999, 0.144428395899013, 0.635050521697849, 0.688108941074461, 0.426051875576377, 0.121159492526203, 0.0143798175267875, 0.755373451858759, 0.698908788152039, 0.575224957661703, 0.485146273626015, 0.272692476864904, 0.64145445660688, 0.936017214087769, 0.316201862413436, 0.856183385709301, 0.539432846242562, 0.236189990770072, 0.800398778636009, 0.869043865473941, 0.566465685144067, 0.634390140417963, 0.882292144466192, 0.122987961629406, 0.570001943036914, 0.275238279951736, 0.305744342040271, 0.651000029174611, 0.466913060983643, 0.137777608120814, 0.270943226059899, 0.87777252192609, 0.0879333338234574, 0.620990377618, 0.0227816230617464, 0.387778277974576, 0.014463109197095, 0.265055202879012, 0.242006823653355, 0.241555956657976, 0.928198719397187, 0.890180074144155, 0.055902763037011, 0.418463099515066, 0.492005074163899, 0.231383620295674, 0.378541557816789, 0.777626163791865, 0.909881828352809, 0.40252801310271, 0.103115998441353, 0.587905461201444, 0.20075240614824, 0.0807150665204972, 0.700731607386842, 0.949362547602504, 0.0912298278417438, 0.464135391870514, 0.427706425311044, 0.368511471198872, 0.117419809103012, 0.227500338805839, 0.949228060664609, 0.316537643782794, 0.279198682168499, 0.487914974568412, 0.0214886313769966, 0.665960716083646, 0.283208351582289, 0.585081983823329, 0.824703778605908, 0.658005223376676, 0.124990732874721, 0.244597244542092, 0.9964377831202, 0.844802460167557, 0.697443865938112, 0.108727243030444, 0.243635117309168, 0.778246119618416, 0.808341949945316, 0.389036037726328, 0.615118770860136, 0.406746171182021, 0.277441482292488, 0.821685605915263, 0.59488990646787, 0.479006396606565, 0.0215427987277508, 0.493229175684974, 0.956676488975063, 0.0160678268875927, 0.121733807493001, 0.75562648079358, 0.323280711658299, 0.99893439351581, 0.92823556647636, 0.137811353197321, 0.79981767386198, 0.149107028031722, 0.16454093134962, 0.0625136815942824, 0.653470316436142, 0.886466546682641, 0.302829560125247, 0.806777003454044, 0.579247435787693, 0.241559724323452, 0.068624904146418, 0.423001967836171, 0.884890978690237, 0.181074557360262, 0.469605535035953, 0.789534627692774, 0.881529270904139, 0.563087606336921, 0.159746304387227, 0.0421771453693509, 0.894298747414723, 0.510756374336779, 0.20665550744161, 0.840002441313118, 0.0951575010549277, 0.936311938799918, 0.668860823381692, 0.894259726163, 0.533416209043935, 0.376375002553686, 0.539821667596698, 0.0281673902645707, 0.436641824431717, 0.0173597454559058, 0.442175193224102, 0.496859249891713, 0.902444201987237, 0.756846515461802, 0.946863359538838, 0.815005607437342, 0.519657118944451, 0.251380074536428, 0.437341197859496, 0.0279886373318732, 0.293678386136889, 0.821313380030915, 0.78529564011842, 0.539934474974871, 0.587724773678929, 0.966229994315654, 0.216423505917192, 0.625733260530978, 0.88996485969983, 0.51424224418588, 0.47671301686205, 0.941978549817577, 0.310940100345761, 0.493798416573554, 0.789315821835771, 0.00777772814035416, 0.353709289804101, 0.142966359388083, 0.775971752824262, 0.46165807521902, 0.830266391392797, 0.222393781645223, 0.723665593424812, 0.845728704473004, 0.233661027392372, 0.0184426154009998, 0.622642762493342, 0.156012231716886, 0.217681041918695, 0.784307497087866, 0.395928285317495, 0.330792339984328, 0.823520925128832, 0.165414900751784, 0.404112940421328, 0.548882298404351, 0.193296993151307, 0.618250552564859, 0.2601089887321, 0.253124615643173, 0.258373975753784, 0.154225663747638, 0.0854834418278188, 0.272937720874324, 0.07394959917292, 0.712043131235987, 0.89360733050853, 0.153152958955616, 0.721404930809513, 0.570232135709375, 0.307491811458021, 0.561548877041787, 0.823958874214441, 0.461753502022475, 0.757299786433578, 0.0345392636954784, 0.163242698414251, 0.901100833434612, 0.769821840338409, 0.34556950815022, 0.834190135356039, 0.207807726925239, 0.260935835540295, 0.684921801788732, 0.672652281122282, 0.0590354811865836, 0.87911263667047, 0.107834935188293, 0.149420205038041, 0.219910636777058, 0.199454542249441, 0.664040799252689, 0.869373768568039, 0.555340353399515, 0.506978470366448, 0.774838972371072, 0.360387979540974, 0.400610349839553, 0.0568663645535707, 0.0722062820568681, 0.101760630263016, 0.690587906865403, 0.907982195960358, 0.0827711580786854, 0.482574781868607, 0.618974441196769, 0.946282600983977, 0.864887986332178, 0.430352613562718, 0.218768067192286, 0.0440493240021169, 0.0917112983297557, 0.704738351050764, 0.229913115268573, 0.0742401222232729, 0.224142149090767, 0.375774920452386, 0.772224326152354, 0.100788897369057, 0.0675555386114866, 0.662708274321631, 0.12336874846369, 0.730346621014178, 0.161740106064826, 0.977253044489771, 0.516488663619384, 0.125402647536248, 0.899686936987564, 0.610091093461961, 0.573825004976243, 0.65460910834372, 0.0144058242440224, 0.99536864971742, 0.293444820446894, 0.0588985960930586, 0.0481104489881545, 0.145386799471453, 0.484262610087171, 0.422943496145308, 0.290970250964165, 0.0116200551856309, 0.0726593227591366, 0.980130066163838, 0.88937799539417, 0.967623896896839, 0.475953418994322, 0.192080973181874, 0.811750683933496, 0.816994874039665, 0.0412322997581214, 0.467329747276381, 0.0222690082155168, 0.130640645278618, 0.883906231960282, 0.761764110997319, 0.977279691491276, 0.365120491944253, 0.728270816616714, 0.124705297872424, 0.058975076302886, 0.272546821506694, 0.746962788514793, 0.952040873002261, 0.342917330097407, 0.717348143225536, 0.456392312655225, 0.419567998033017, 0.867686879821122, 0.644594149664044, 0.446037546498701, 0.767598482081667, 0.496296271216124, 0.168802331667393, 0.927406842354685, 0.303985730046406, 0.600081009557471, 0.0814471687190235, 0.201651441166177, 0.146651254966855, 0.175355486106128, 0.858178687747568, 0.210027845576406, 0.341279343236238, 0.417882041772828, 0.00684744957834482, 0.182296226499602, 0.468382864026353, 0.345362401334569, 0.0442152128089219, 0.36951502552256, 0.729080657474697, 0.0420440854504704, 0.198373604565859, 0.502874172758311, 0.96697179810144, 0.544676476623863, 0.803933015093207, 0.920770476106554, 0.0566882933489978, 0.654051890363917, 0.348495612852275, 0.200647981371731, 0.540632475633174, 0.537404115311801, 0.0475928545929492, 0.451445433078334, 0.505524240667, 0.524808002868667, 0.94588367943652, 0.166456710081547, 0.97578246309422, 0.148501765914261, 0.373679417883977, 0.494053381960839, 0.919295186176896, 0.716077933320776, 0.367306776344776, 0.141754085896537, 0.794311652192846, 0.805160818621516, 0.66255604615435, 0.118763432605192, 0.112806935794652, 0.0864354178775102, 0.620244936319068, 0.0212281008716673, 0.879105515545234, 0.774122908245772, 0.97658642847091, 0.610520224319771, 0.219174990430474, 0.134346334962174, 0.282367894658819, 0.717740157619119, 0.00595858437009156, 0.664318474940956, 0.580141810234636, 0.924371077911928, 0.766476991120726, 0.908692089142278, 0.0759362645912915, 0.515837421175092, 0.595595226855949, 0.00587678537704051, 0.260488356696442, 0.0320140968542546, 0.0798783239442855, 0.863567026564851, 0.982973825419322, 0.434193204157054, 0.725188278360292, 0.746022567385808, 0.673402769258246, 0.807090119458735, 0.836209211032838, 0.106188735226169, 0.337993773864582, 0.99153918097727, 0.0887104079592973, 0.455889678560197, 0.206823926186189, 0.96564602619037, 0.358317917212844, 0.808659745380282, 0.588041879003868, 0.465123304864392, 0.963223601458594, 0.516178985591978, 0.846233271528035, 0.948945916257799, 0.967098807683215, 0.619824137771502, 0.489919743034989, 0.969849013490602, 0.971931028645486, 0.588116439990699, 0.752119518816471, 0.26967816776596, 0.184893439058214, 0.655461406335235, 0.831070423126221, 0.0792892659083009, 0.747918607899919, 0.881358959944919, 0.383398961042985, 0.186772209592164, 0.63336154865101, 0.600530863506719, 0.376169701805338, 0.399787329370156, 0.427228869870305, 0.168524210806936, 0.798656758386642, 0.532401840435341, 0.809645785484463, 0.760439200792462, 0.10991327627562, 0.0926491639111191, 0.422885082894936, 0.0149017537478358, 0.595222290372476, 0.719665681244805, 0.72289804648608, 0.909457941306755, 0.323077751789242, 0.344392167869955, 0.336469057947397, 0.576839397428557, 0.683475844329223, 0.206490429583937, 0.124186662724242, 0.791336456779391, 0.866047007963061, 0.0494966271799058, 0.480372012360021, 0.0643995604477823, 0.732655055820942, 0.0639734263531864, 0.159736109431833, 0.596876805415377, 0.73406033217907, 0.829097290989012, 0.827940376009792, 0.913780942792073, 0.572497212793678, 0.885168651351705, 0.314791293581948, 0.317945756250992, 0.8321015490219, 0.496840495849028, 0.42571695940569, 0.143003979232162, 0.604324796935543, 0.753554561408237, 0.488654457032681, 0.0500525834504515, 0.101495772367343, 0.733974842820317, 0.369652190245688, 0.462706982390955, 0.115871608257294, 0.138777903281152, 0.87093352759257, 0.0823463043197989, 0.707211216911674, 0.186675965320319, 0.0445223650895059, 0.0466795682441443, 0.185488964896649, 0.258265204261988, 0.287704500136897, 0.947532356251031, 0.0386092849075794, 0.771249936660752, 0.840265526901931, 0.901421374175698, 0.623429967556149, 0.706770052900538, 0.256203683558851, 0.259313917485997, 0.514762507518753, 0.679639253299683, 0.731831816257909, 0.322172719519585, 0.418032511603087, 0.794309665681794, 0.440625685499981, 0.00497913872823119, 0.0967818051576614, 0.643459621816874, 0.620771375251934, 0.215390485944226, 0.775357985636219, 0.469319554045796, 0.144533465383574, 0.377335339551792, 0.442695825127885, 0.180874639656395, 0.948987489799038, 0.301698708673939, 0.853861261857674, 0.0108477054163814, 0.486265784827992, 0.456134040374309, 0.447858022758737, 0.295277352211997, 0.00708449189551175, 0.308356846449897, 0.176792735466734, 0.227252054261044, 0.359545757994056, 0.380382267991081, 0.823074916144833, 0.316874156007543, 0.204791673226282, 0.56804153858684, 0.490960309281945, 0.537255932111293, 0.703178780153394, 0.631680994760245, 0.42559610423632, 0.494670590618625, 0.864390667993575, 0.818313755095005, 0.143888369901106, 0.793619524454698, 0.0395210317801684, 0.0699483617208898, 0.37157005793415, 0.306421338813379, 0.858017069287598, 0.775690580252558, 0.0878839420620352, 0.832015090389177, 0.893482076004148, 0.700347885256633, 0.308705136179924, 0.387677824590355, 0.877884045941755, 0.709198790369555, 0.654803355224431, 0.574209883110598, 0.810215517412871, 0.496146718040109, 0.0376430212054402, 0.486685349838808, 0.817312049679458, 0.937628591200337, 0.875082820188254, 0.665190093452111, 0.17856284067966, 0.0669604362919927, 0.902600183384493, 0.832999418722466, 0.475042635807768, 0.386113481596112, 0.249234412331134, 0.707041377900168, 0.805946786887944, 0.866969665046781, 0.167941942345351, 0.468656752957031, 0.754063718952239, 0.493156412150711, 0.467059957096353, 0.719051119405776, 0.950662165647373, 0.387234682450071, 0.618803601479158, 0.598696274682879, 0.366512110689655, 0.859318868489936, 0.242307558655739, 0.100632338784635, 0.699582972330973, 0.650293975835666, 0.107657727319747, 0.944799266755581, 0.24272150429897, 0.112025308422744, 0.755802327534184, 0.20287726027891, 0.201090027578175, 0.78104130201973, 0.151889879489318, 0.901604109210894, 0.827701388159767, 0.845446795225143, 0.168880825862288, 0.763947244267911, 0.729779136134312, 0.749443640001118, 0.170062275137752, 0.395876705413684, 0.608244190691039, 0.598776045488194, 0.747517231153324, 0.797019256278872, 0.202969153877348, 0.522406484000385, 0.371475564548746, 0.615015007322654, 0.561170019442216, 0.224414320895448, 0.118199829477817, 0.529995770426467, 0.0827378949616104, 0.169686908600852, 0.941142118768767, 0.77256844798103, 0.273240311769769, 0.320573304314166, 0.839052881114185, 0.46349483076483, 0.200430443976074, 0.236742556095123, 0.451133044669405, 0.558713666861877, 0.891182596795261, 0.108650414505973, 0.38629982015118, 0.942521943477914, 0.0708604760002345, 0.981139211216941, 0.239757322473451, 0.448984909104183, 0.0976594961248338, 0.721426518866792, 0.87420678883791, 0.919751660665497, 0.631085597211495, 0.773347465088591, 0.920360272051767, 0.285187749424949, 0.911554157035425, 0.228410569019616, 0.243091899901628, 0.781793379224837, 0.685098007088527, 0.32608751882799, 0.162569212727249, 0.486288443207741, 0.044459130614996, 0.104780968511477, 0.460313169984147, 0.556270273635164, 0.0082236984744668, 0.832087922841311, 0.787622372386977, 0.935591791989282, 0.74771926109679, 0.43050278397277, 0.193486862815917, 0.436673447024077, 0.96233924292028, 0.333046514540911, 0.918753575067967, 0.276206807233393, 0.117944598896429, 0.919833352090791, 0.881605005823076, 0.275027761468664, 0.397776020690799, 0.728989048162475, 0.870054366765544, 0.221888123080134, 0.900296325096861, 0.959091408876702, 0.222617820370942, 0.2790307700634, 0.209341198438779, 0.860053269891068, 0.578009312273934, 0.960597649216652, 0.211155690951273, 0.287912003928795, 0.119710033293813, 0.28898416692391, 0.442746093962342, 0.524869766319171, 0.194827504921705, 0.74533047247678, 0.0895033406559378, 0.189436384476721, 0.678182317409664, 0.01324657863006, 0.999089299002662, 0.441376587375998, 0.251589567633346, 0.648543497314677, 0.926471152110025, 0.787135580321774, 0.848129295744002, 0.383148608263582, 0.117224272806197, 0.109839149052277, 0.225745664676651, 0.0405693859793246, 0.25780336977914, 0.760293721221387, 0.552792397560552, 0.329479621024802, 0.683489573886618, 0.68944536568597, 0.802757585188374, 0.463085143361241, 0.354483491741121, 0.566696971887723, 0.690133238676935, 0.50459124147892, 0.424474883358926, 0.252165094949305, 0.266988345887512, 0.667231161147356, 0.674684632569551, 0.723616200499237, 0.401376914698631, 0.797468570992351, 0.296735184732825, 0.287298901472241, 0.108868307201192, 0.392998773837462, 0.747177035547793, 0.530504071852192, 0.649255495751277, 0.681850242894143, 0.0592208479065448, 0.235835063504055, 0.742773894453421, 0.237415771232918, 0.696322173345834, 0.706383321201429, 0.722661613952368, 0.823639226844534, 0.207200655713677, 0.394033688120544, 0.290291990153491, 0.0996042324695736, 0.509365360019729, 0.845544703304768, 0.484104128088802, 0.497102913912386, 0.636119060218334, 0.426348131382838, 0.305868286173791, 0.0139504529070109, 0.196925349300727, 0.614866912830621, 0.955936820711941, 0.751545884413645, 0.771032779011875, 0.678424939280376, 0.204760113731027, 0.192871318897232, 0.232216594042256, 0.545906120445579, 0.629490786464885, 0.477910266024992, 0.941575862001628, 0.446324384538457, 0.984257386764511, 0.194033333333209, 0.308152961544693, 0.976371842669323, 0.989842681447044, 0.643851994071156, 0.716555525083095, 0.599204070633277, 0.968065333552659, 0.495976972160861, 0.376440640538931, 0.596486969152465, 0.913929746020585, 0.025144113227725, 0.236525814514607, 0.712210881523788, 0.895865958649665, 0.664717372739688, 0.838343260111287, 0.687969743041322, 0.171786425868049, 0.0635133720934391, 0.722244848497212, 0.979349037166685, 0.48431152314879, 0.454374566674232, 0.0732750205788761, 0.229843975044787, 0.324924984714016, 0.0679655421990901, 0.030429188394919, 0.545595472212881, 0.437323480378836, 0.687547917710617, 0.0681631350889802, 0.871673917165026, 0.544949056813493, 0.354371556546539, 0.322755991248414, 0.214122185250744, 0.339414578862488, 0.927710606250912, 0.0477268453687429, 0.511976813897491, 0.790436537005007, 0.232822557212785, 0.670060378033668, 0.0925950098317116, 0.306428539566696, 0.00207580113783479, 0.37501116655767, 0.405283309519291, 0.219594713300467, 0.00141216907650232, 0.671518712770194, 0.585365517996252, 0.585767831187695, 0.0662378252018243, 0.260477766860276, 0.806606637779623, 0.275564266135916, 0.801295862300321, 0.349520154763013, 0.185863004298881, 0.755709919612855, 0.780877156881616, 0.319992039818317, 0.0704637942835689, 0.64502025465481, 0.839841727167368, 0.58142819092609, 0.412513874936849, 0.0225519365631044, 0.704829484224319, 0.567530597560108, 0.245952994329855, 0.539494127267972, 0.588333410909399, 0.318400850985199, 0.5508651684504, 0.126217819750309, 0.528532752301544, 0.987473109969869, 0.939408123726025, 0.339521691203117, 0.600431634346023, 0.167609413154423, 0.436033786274493, 0.0769396538380533, 0.280445974785835, 0.122723655076697, 0.877063601976261, 0.522419218206778, 0.358914355747402, 0.938245440134779, 0.280957065522671, 0.188423527404666, 0.767907740082592, 0.450512368232012, 0.673774952534586, 0.629311869153753, 0.445811427664012, 0.475375400856137, 0.155310734407976, 0.0131323528476059, 0.476971727795899, 0.638769775163382, 0.531940004089847, 0.828015372157097, 0.075822509592399, 0.354146058671176, 0.302415763493627, 0.228245040634647, 0.298650417476892, 0.00120612513273954, 0.132904721889645, 0.520200223661959, 0.337588076014072, 0.930435674963519, 0.769251006888226, 0.61503145727329, 0.501837157411501, 0.66468867030926, 0.183618466136977, 0.0722061244305223, 0.994687279919162, 0.79849541047588, 0.489734278991818, 0.442391215125099, 0.984176956117153, 0.197635160991922, 0.857497981749475, 0.986956618726254, 0.231289088726044, 0.475474445382133, 0.942332861246541, 0.942381409462541, 0.221137353917584, 0.558749303920195, 0.172455512685701, 0.846294205868617, 0.0823663354385644, 0.385989002417773, 0.583416827721521, 0.746658152900636, 0.185757798142731, 0.379119223449379, 0.480142802465707, 0.441917538875714, 0.351548282196745, 0.669829024700448, 0.810923439916223, 0.155911799753085, 0.0731936513911933, 0.0722583844326437, 0.600200805347413, 0.469359399517998, 0.974019431509078, 0.538100532954559, 0.757666458142921, 0.761853673961014, 0.982679309789091, 0.115523232147098, 0.399608808336779, 0.647438112879172, 0.0680333795025945, 0.552275758236647, 0.850604565348476, 0.187629019841552, 0.825552959926426, 0.208034641342238, 0.694130546879023, 0.00321755115874112, 0.437356385402381, 0.384709779405966, 0.551621970720589, 0.968065183144063, 0.436183177633211, 0.234541383106261, 0.162049649748951, 0.933063177159056, 0.068867702037096, 0.168855714378878, 0.724074597470462, 0.768371772952378, 0.13351227925159, 0.486117051914334, 0.468513805652037, 0.822844080626965, 0.748936872696504, 0.395217955345288, 0.500948552275077, 0.210264868335798, 0.366946632275358, 0.437091644853354, 0.991386799141765, 0.7240939817857, 0.645571268396452, 0.34134744759649, 0.33440457470715, 0.560506860492751, 0.934905784437433, 0.600012312410399, 0.0555973497685045, 0.633055120939389, 0.680280688451603, 0.88982002902776, 0.148392303148285, 0.495836276095361, 0.954519863706082, 0.955867143115029, 0.759199839318171, 0.278576399199665, 0.724936632206663, 0.096071318956092, 0.694588870042935, 0.311899404739961, 0.757997731445357), actual = c(0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.00969696557149291, 0.786896218080074, 0.955669911112636, 0.901219836901873, 0.444982967805117, 0.199930173577741, 0.299359677359462, 0.126153642777354, 0.665084214415401, 0.182705185143277, 0.0741628929972649, 0.489107970148325, 0.27170882653445, 0.709780963836238, 0.763658702373505, 0.463955208659172, 0.750311407726258, 0.175643590278924, 0.558350087376311, 0.971229162067175, 0.822819229215384, 0.504048770526424, 0.924413718050346, 0.842868561157957, 0.478105742717162, 0.750795798841864, 0.988811626564711, 0.386232384247705, 0.729320199694484, 0.207902817754075, 0.643057423178107, 0.145932018058375, 0.105146032758057, 0.417441639350727, 0.750605543842539, 0.985733563313261, 0.890967151150107, 0.394091191701591, 0.418216135120019, 0.661876192782074, 0.637934408150613, 0.512580899056047, 0.998172723222524, 0.192432290641591, 0.892510398291051, 0.341866538859904, 0.838144759181887, 0.926098734838888, 0.579169884789735, 0.600928779458627, 0.682563992217183, 0.552837244933471, 0.591438965406269, 0.273496634094045, 0.102991605876014, 0.0977821792475879, 0.116470016539097, 0.351816945010796, 0.583523228531703, 0.226156081072986, 0.109830693807453, 0.98425803356804, 0.188745919149369, 0.958980136085302, 0.06181844137609, 0.798797016497701, 0.315465965308249, 0.792332495795563, 0.579001974314451, 0.465562998550013, 0.759008365217596, 0.299050633562729, 0.96643606829457, 0.0109691524412483, 0.422382242511958, 0.969426108524203, 0.79974625329487, 0.12735723820515, 0.52915409184061, 0.854369560489431, 0.69292117189616, 0.78228975366801, 0.423525273334235, 0.930691667366773, 0.826811277074739, 0.878307102713734, 0.458097668830305, 0.0145650263875723, 0.0898559961933643, 0.611781604355201, 0.153104925062507, 0.917812940664589, 0.329638093011454, 0.680458747316152, 0.0559555867221206, 0.49035494425334, 0.847731756977737, 0.456090637017041, 0.947244287002832, 0.116294459672645, 0.493647880153731, 0.176936213159934, 0.341419223695993, 0.715187536785379, 0.88235294027254, 0.120139937382191, 0.819466522661969, 0.721620843512937, 0.59859633166343, 0.873020062921569, 0.676352944923565, 0.407607653411105, 0.730059979017824, 0.526817084988579, 0.53026794269681, 0.0719578759744763, 0.603204035433009, 0.39339443622157, 0.240005372557789, 0.151414327323437, 0.310134983155876, 0.550962222972885, 0.540068897651508, 0.62526249117218, 0.920816012658179, 0.156462069135159, 0.254737768787891, 0.33110528672114, 0.416722487425432, 0.308025236940011, 0.836837539449334, 0.675535806454718, 0.792349639581516, 0.824202479794621, 0.020134539110586, 0.268738884711638, 0.122871668310836, 0.0557421394623816, 0.285614407388493, 0.342053624102846, 0.957811006344855, 0.459552376531065, 0.502685192041099, 0.0892373942770064, 0.241374290781096, 0.308371950406581, 0.758186024846509, 0.613255896372721, 0.0161155564710498, 0.0804533022455871, 0.941527884453535, 0.914158585248515, 0.0239096749573946, 0.797602242324501, 0.084944105707109, 0.287336535518989, 0.771639478160068, 0.393370918463916, 0.436553397448733, 0.762733461568132, 0.525391859700903, 0.805149498861283, 0.794251281768084, 0.437327492516488, 0.486077927285805, 0.450183984357864, 0.947410011896864, 0.576223448617384, 0.585101035889238, 0.0915793208405375, 0.65435112756677, 0.352974115405232, 0.883314572740346, 0.571410780074075, 0.0882322266697884, 0.996993651846424, 0.145734803983942, 0.368893294362351, 0.841232289327309, 0.0538992653600872, 0.811988953035325, 0.273406758671626, 0.678686198312789, 0.611234464682639, 0.226671527605504, 0.67629854590632, 0.428854467114434, 0.456867975415662, 0.126321142772213, 0.877458897186443, 0.403590492904186, 0.411255159880966, 0.398939886130393, 0.969279618468136, 0.203634703997523, 0.708566936431453, 0.503934722859412, 0.557361277285963, 0.357652283739299, 0.135360880987719, 0.00565951922908425, 0.571857115952298, 0.333361971657723, 0.731450476683676, 0.746889285510406, 0.976510297274217, 0.39665216486901, 0.244534574681893, 0.428534065140411, 0.595720619196072, 0.560050670988858, 0.802874394925311, 0.270311869913712, 0.444004886317998, 0.306866412749514, 0.645294947782531, 0.0990309233311564, 0.294834333239123, 0.50057152309455, 0.3428272055462, 0.294307773699984, 0.3006332507357, 0.721412430983037, 0.823542492697015, 0.189813094679266, 0.398169188760221, 0.68518063868396, 0.731060364050791, 0.127794187748805, 0.917164469137788, 0.120720692677423, 0.595976673997939, 0.928751569474116, 0.369719399139285, 0.223559289937839, 0.410696957493201, 0.60384618723765, 0.237305442569777, 0.968144924379885, 0.658429219387472, 0.791691086255014, 0.964689614716917, 0.39015718922019, 0.342333118896931, 0.795169465709478, 0.679188612382859, 0.431523218750954, 0.216419717064127, 0.511131246807054, 0.404416785342619, 0.0364361484535038, 0.539126418763772, 0.613346446771175, 0.47659284924157, 0.713839802891016, 0.958975315792486, 0.725546597968787, 0.483438260853291, 0.502335227327421, 0.402130578644574, 0.226725112879649, 0.0329949555452913, 0.690548033919185, 0.755712160374969, 0.986050980864093, 0.951651320094243, 0.684156082803383, 0.0191017908509821, 0.87008881336078, 0.647946414304897, 0.559429613640532, 0.0637760374229401, 0.173141334904358, 0.532186555443332, 0.307185347191989, 0.305199097609147, 0.806484400294721, 0.651805372908711, 0.23084095842205, 0.218286422314122, 0.982195354765281, 0.378864411497489, 0.68839430087246, 0.564146007411182, 0.358406495768577, 0.105068209115416, 0.647741730092093, 0.343071826035157, 0.0601132723968476, 0.768261287361383, 0.0212153026368469, 0.394044454675168, 0.0331424965988845, 0.793891185894608, 0.623239813139662, 0.0224950581323355, 0.135326568270102, 0.0632913007866591, 0.502588080940768, 0.645302967168391, 0.628942406270653, 0.54995113844052, 0.214437103131786, 0.311687720473856, 0.389601699076593, 0.0314596735406667, 0.273682402214035, 0.667686140863225, 0.865261361002922, 0.127132409485057, 0.647028404055163, 0.409211458172649, 0.821883193682879, 0.672634845133871, 0.0759273751173168, 0.050431665731594, 0.213769792811945, 0.621809522621334, 0.507481985725462, 0.467788859270513, 0.840576839400455, 0.880520232953131, 0.454525899142027, 0.0654819773044437, 0.0397365565877408, 0.562334051821381, 0.118974360171705, 0.553764802869409, 0.862843500217423, 0.967630452942103, 0.361366923665628, 0.33007645024918, 0.986509338254109, 0.655470489058644, 0.280134449247271, 0.535355001688004, 0.947442254051566, 0.32717058993876, 0.951385163702071, 0.362782133277506, 0.405262144980952, 0.303352914983407, 0.192930231336504, 0.337442638352513, 0.738251625327393, 0.679938508430496, 0.459387624170631, 0.981026034103706, 0.0161485637072474, 0.860803722636774, 0.899009973974898, 0.955041106790304, 0.0329189612530172, 0.412634710548446, 0.186955283628777, 0.954391581937671, 0.814488331787288, 0.0305038644000888, 0.984410892706364, 0.28374916058965, 0.0512760581914335, 0.965469812741503, 0.786474812310189, 0.417983786901459, 0.191264342516661, 0.0806666924618185, 0.238424868090078, 0.869753939798102, 0.708484996808693, 0.0457045219372958, 0.605054588522762, 0.166356798727065, 0.274762040004134, 0.431809149216861, 0.378672615392134, 0.502402040176094, 0.604022122686729, 0.641780353849754, 0.162656743312255, 0.517619383288547, 0.202360669150949, 0.389868923230097, 0.976294995285571, 0.781242204131559, 0.10900416597724, 0.564828027738258, 0.950483930530027, 0.258117317454889, 0.130332796135917, 0.0993458477314562, 0.00390121480450034, 0.104598788777366, 0.522343845805153, 0.661482339259237, 0.640230594901368, 0.617183567257598, 0.574004352325574, 0.252134166192263, 0.926419677678496, 0.106411136453971, 0.393294823821634, 0.71113887662068, 0.0867344767320901, 0.173410937888548, 0.458909646607935, 0.0909859980456531, 0.336781201884151, 0.954427270684391, 0.584183493396267, 0.672530075535178, 0.180256122490391, 0.788716710405424, 0.0317980677355081, 0.65966596850194, 0.703424453502521, 0.750566536793485, 0.145440221764147, 0.169876243686303, 0.381723126862198, 0.51042016572319, 0.698625488206744, 0.823546845233068, 0.286019468447194, 0.743358867708594, 0.0422985015902668, 0.251429279800504, 0.230060425121337, 0.804771429626271, 0.340378079097718, 0.855326484655961, 0.214514235500246, 0.563818784197792, 0.0292739307042211, 0.104612556053326, 0.604531297693029, 0.915723114274442, 0.319572123931721, 0.5112334783189, 0.298626642208546, 0.279028503224254, 0.704131037229672, 0.679177704500034, 0.868216797010973, 0.799507146701217, 0.639039718313143, 0.0895062661729753, 0.661405030172318, 0.942441211082041, 0.306886625243351, 0.361020581563935, 0.989404572639614, 0.376778283622116, 0.809020965592936, 0.666802832623944, 0.205049734329805, 0.0307253261562437, 0.915965748019516, 0.790221486007795, 0.663607449270785, 0.0455687297508121, 0.465970285935327, 0.303223506780341, 0.52832327154465, 0.960773115279153, 0.914037566166371, 0.364711920265108, 0.0870891229715198, 0.547705068252981, 0.744483261136338, 0.460937622003257, 0.447908327449113, 0.51871082582511, 0.761955376481637, 0.815428150584921, 0.781110580777749, 0.659931864123791, 0.0680712794419378, 0.996388669591397, 0.42315112403594, 0.0235801166854799, 0.18863346031867, 0.390029860194772, 0.249464209889993, 0.933280159486458, 0.316894013667479, 0.0834022436756641, 0.940065032336861, 0.053696209564805, 0.516576788155362, 0.336234274553135, 0.988744375295937, 0.575253817718476, 0.886396958027035, 0.908387974370271, 0.953985739033669, 0.0117089268751442, 0.719439923064783, 0.904627114767209, 0.284518625587225, 0.630160792265087, 0.287056126864627, 0.877657813485712, 0.308615154121071, 0.196797380689532, 0.493810905143619, 0.752456812886521, 0.180553659796715, 0.732451686635613, 0.44339603302069, 0.312759287422523, 0.371915379539132, 0.399244620930403, 0.441956504480913, 0.511080211261287, 0.580920168897137, 0.890638354467228, 0.300739440135658, 0.251802251674235, 0.801914446055889, 0.745625479379669, 0.792598115745932, 0.984094536630437, 0.440125143621117, 0.538959300145507, 0.277644671499729, 0.87402412854135, 0.00656617851927876, 0.878233555704355, 0.917116709519178, 0.289721686625853, 0.284590182825923, 0.536455553490669, 0.237908328417689, 0.981799459317699, 0.994540128624067, 0.960585480090231, 0.748897843295708, 0.85098275821656, 0.81488216877915, 0.64589544897899, 0.314638163894415, 0.594958262518048, 0.585537781473249, 0.783969044918194, 0.397369406884536, 0.744580170605332, 0.431156989419833, 0.772707058116794, 0.938807633472607, 0.925749334041029, 0.0761021003127098, 0.123570932773873, 0.0500775438267738, 0.292349686380476, 0.17678994894959, 0.892074313247576, 0.25681797740981, 0.142318267840892, 0.752056771889329, 0.479919749312103, 0.231760815251619, 0.114165397826582, 0.713743116473779, 0.492629769956693, 0.688429071800783, 0.654413357609883, 0.476017254171893, 0.754127643303946, 0.370635336264968, 0.383346058893949, 0.658783731516451, 0.427011785330251, 0.402171661145985, 0.62195334257558, 0.661042566876858, 0.369197271997109, 0.285225961357355, 0.833797816885635, 0.700466223759577, 0.197637628763914, 0.239312016172335, 0.513159193331376, 0.79597517545335, 0.582001872826368, 0.467683810042217, 0.741827391786501, 0.23812002572231, 0.840589145664126, 0.204175965627655, 0.0393392860423774, 0.857934701954946, 0.251357235014439, 0.817903378978372, 0.765788058517501, 0.586829250212759, 0.234724243171513, 0.429688756819814, 0.21007507503964, 0.370621394133195, 0.378205663291737, 0.77013403060846, 0.565127481007949, 0.632992681581527, 0.53435174911283, 0.882668342674151, 0.0939794324804097, 0.0226872358471155, 0.653485372662544, 0.749526774743572, 0.480908387340605, 0.186403883853927, 0.764954924117774, 0.753054769476876, 0.963200806174427, 0.849002862581983, 0.611096285050735, 0.532987294718623, 0.86022500321269, 0.487226702738553, 0.818595674354583, 0.213853339664638, 0.93365229992196, 0.977187578333542, 0.813904042821378, 0.811699258396402, 0.498714491026476, 0.98695281217806, 0.217904596822336, 0.01365732611157, 0.767433985136449, 0.960409868741408, 0.0220516007393599, 0.507777463411912, 0.33221953199245, 0.0656257856171578, 0.396537006599829, 0.291071087354794, 0.374516900628805, 0.89920192072168, 0.585147245321423, 0.359572150046006, 0.356658798642457, 0.765264017041773, 0.0272523353341967, 0.148957914672792, 0.441602657549083, 0.22973300749436, 0.0871284119784832, 0.402579609770328, 0.479635769734159, 0.028609424829483, 0.911645675078034, 0.0167777389287949, 0.541936716763303, 0.32775641954504, 0.229083174839616, 0.893769814632833, 0.0215017846785486, 0.443485648836941, 0.466551403515041, 0.783883555326611, 0.0613936576992273, 0.386963615426794, 0.0945776544976979, 0.886589254718274, 0.121355682378635, 0.547240392072126, 0.783752585062757, 0.873938862001523, 0.94681115122512, 0.406134897610173, 0.7319881611038, 0.68332694331184, 0.962840860243887, 0.392083427868783, 0.467357067624107, 0.0887216918636113, 0.804316350724548, 0.864294937811792, 0.29057314991951, 0.509780487278476, 0.987058246508241, 0.742713919375092, 0.735602609580383, 0.161855302751064, 0.795480972155929, 0.176013414282352, 0.394551450153813, 0.803568632341921, 0.216980150900781, 0.608150431420654, 0.0356761997099966, 0.705368014751002, 0.776644620345905, 0.913531325291842, 0.65292727551423, 0.100893932161853, 0.417965880595148, 0.49633052572608, 0.189126040786505, 0.375162317184731, 0.404630580451339, 0.0627978043630719, 0.218237716704607, 0.551336839096621, 0.235579638276249, 0.169639897998422, 0.970929382834584, 0.987028601812199, 0.578615411417559, 0.868270041886717, 0.894888281822205, 0.018288365798071, 0.302652849582955, 0.605188658460975, 0.769577156519517, 0.407299050595611, 0.240147163392976, 0.277441994985566, 0.55654246173799, 0.416607799008489, 0.230924144387245, 0.257857864489779, 0.212024087784812, 0.602867326000705, 0.387849061284214, 0.554286828264594, 0.212170449085534, 0.639061388559639, 0.163805705728009, 0.949407340725884, 0.192095800535753, 0.981302271829918, 0.0835530017502606, 0.0601530822459608, 0.67670399043709, 0.850826977053657, 0.392714489484206, 0.361080603208393, 0.0284883864223957, 0.856863487977535, 0.0715083056129515, 0.308213300071657, 0.937985678436235, 0.445763774914667, 0.301874975906685, 0.172125302487984, 0.462851449847221, 0.439205617643893, 0.411498725879937, 0.845657263882458, 0.816037691896781, 0.587962145917118, 0.544807304628193, 0.821432350436226, 0.775668420130387, 0.841845758259296, 0.592089045559987, 0.159736791159958, 0.202732819830999, 0.366035138489679, 0.727536529535428, 0.998388511594385, 0.635345939779654, 0.0879259316716343, 0.774061163654551, 0.48865313641727, 0.408724444918334, 0.652238762704656, 0.139402962522581, 0.970342872897163, 0.0533943644259125, 0.601453670999035, 0.373043033760041, 0.502517635235563, 0.848329812986776, 0.032657200936228, 0.619020616868511, 0.0461497663054615, 0.756022183923051, 0.114272254053503, 0.0119975875131786, 0.480045659933239, 0.313422965817153, 0.546293838648126, 0.30871150130406, 0.186763560399413, 0.519787410041317, 0.316911776084453, 0.746562773594633, 0.832599449902773, 0.304002330638468, 0.952008004300296, 0.361706476658583, 0.461332138162106, 0.0302117008250207, 0.762172990944237, 0.915346996858716, 0.987290563993156, 0.600184860872105, 0.68789850990288, 0.983086601132527, 0.987479719333351, 0.691291695926338, 0.31932458281517, 0.494724726537243, 0.994096166687086, 0.395333416759968, 0.555711161112413, 0.36384428688325, 0.846386488992721, 0.563386331079528, 0.398172978078946, 0.962810881901532, 0.313295325264335, 0.978239080635831, 0.261109696235508, 0.408154472708702, 0.555345780216157, 0.197824902832508, 0.897693815175444, 0.579750745557249, 0.333460585447028, 0.939215203979984, 0.547244918998331, 0.000753246014937758, 0.228227708023041, 0.259341622469947, 0.720527339493856, 0.25487688370049, 0.641775517957285, 0.811860549962148, 0.134518516482785, 0.119559220038354, 0.800132111646235, 0.325406509684399, 0.113255236530676, 0.157961432123557, 0.132452177582309, 0.96381082595326, 0.871275029610842, 0.640396462287754, 0.371220496250317, 0.570591271156445, 0.753887899685651, 0.562423853436485, 0.196923536481336, 0.027583105256781, 0.954063937533647, 0.85151386493817, 0.0181092426646501, 0.944136697798967, 0.113927181810141, 0.0570521678309888, 0.984951641177759, 0.021507635479793, 0.97518819780089, 0.598802202614024, 0.0645870936568826, 0.132021562196314, 0.572631328133866, 0.718311980366707, 0.289724388625473, 0.932139229495078, 0.338958062697202, 0.365221292711794, 0.0971202768851072, 0.648685718886554, 0.168673474341631, 0.295794177101925, 0.162609741091728, 0.360182920470834, 0.317249220563099, 0.62507006758824, 0.206602134276181, 0.402240806724876, 0.409436522284523, 0.975992947118357, 0.530711002647877, 0.72511686058715, 0.991351481061429, 0.700480268802494, 0.848969340324402, 0.402456661453471, 0.339442370226607, 0.81733451411128, 0.174028627574444, 0.588441882282495, 0.256493522785604, 0.148892936995253, 0.404611015226692, 0.72527304943651, 0.562168525997549, 0.725943842669949, 0.318188189063221, 0.929166219197214, 0.530310597503558, 0.64075885550119, 0.047767473384738, 0.425723302643746, 0.542878486448899, 0.679041345138103, 0.94437850615941, 0.167042778804898, 0.479646949563175, 0.213353521190584, 0.139186548767611, 0.270278026349843, 0.117536251898855, 0.102359104901552, 0.243145700078458, 0.337911148322746, 0.381453992100433, 0.0568388581741601, 0.738343331264332, 0.514549993211403, 0.600657780654728, 0.942453449126333, 0.80686245393008, 0.0187813192605972, 0.611379116540775, 0.750307639362291, 0.755900147138163, 0.96165168681182, 0.206009466201067, 0.624920551665127, 0.418069618055597, 0.571035244036466, 0.104560117470101, 0.360380716854706, 0.106187057914212, 0.193283241940662, 0.862339048180729, 0.477638165699318, 0.223276706412435, 0.772982564754784, 0.489451343892142, 0.472830640152097, 0.706367350416258, 0.41823921026662, 0.242411290528253, 0.257653046632186, 0.253555175615475, 0.168518485501409, 0.30361499497667, 0.504966023610905, 0.0987649499438703, 0.824908436043188, 0.352106248261407, 0.265514815924689, 0.954954535700381, 0.6516598330345, 0.237434590701014, 0.640510332072154, 0.30046204617247, 0.649262483697385, 0.719380348920822, 0.964942344930023, 0.651422017952427, 0.440020185895264, 0.644541039830074, 0.710723722120747, 0.348852131748572, 0.930128572043031, 0.0240265575703233, 0.0201451587490737, 0.269137379480526, 0.650834635132924, 0.726295528234914, 0.924591155955568, 0.090595030458644, 0.508503040764481, 0.660910086007789, 0.291964770527557, 0.425794968381524, 0.339777910616249, 0.248024324188009, 0.578109478810802, 0.847209483617917, 0.704814616590738, 0.647934614447877, 0.342181778047234, 0.0656001754105091, 0.686505809891969, 0.914315387839451, 0.975370406638831, 0.539902729913592, 0.253472704207525, 0.268278610426933, 0.387373430887237, 0.69703669170849, 0.906512258108705, 0.371985253877938, 0.131142499856651, 0.229426795616746, 0.696429562056437, 0.46790484781377, 0.886609668610618, 0.835942178731784, 0.782760222209617), actual = c(0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"