Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.9886354        0.7783656 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5062156        0.5062156        0.4899655 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.197485173353925, 0.670890052337199, 0.624194976175204, 0.784243857953697, 0.457691313000396, 0.85992314433679, 0.416358083719388, 0.963161475490779, 0.0216375307645649, 0.723653727676719, 0.709409933770075, 0.650343950372189, 0.500964037608355, 0.957459822297096, 0.254424857906997, 0.923507270868868, 0.72644108068198, 0.458189068129286, 0.374802141217515, 0.859139167703688, 0.605945423711091, 0.234752885298803, 0.783039606176317, 0.101117160404101, 0.554603781318292, 0.28268946451135, 0.165135241346434, 0.684442611411214, 0.974586771568283, 0.0255172250326723, 0.28646371839568, 0.176011758157983, 0.321049820864573, 0.152474504429847, 0.215438042068854, 0.823230820009485, 0.761485895840451, 0.329605216393247, 0.620234008878469, 0.0249811047688127, 0.710559537401423, 0.758592886850238, 0.513036226388067, 0.843525421107188, 0.741172952577472, 0.409102252917364, 0.276668723672628, 0.910891406005248, 0.662520959740505, 0.804016449255869, 0.766437386162579, 0.0791083155199885, 0.460763976443559, 0.865752814803272, 0.779115065233782, 0.725441585993394, 0.0183145634364337, 0.683903520461172, 0.196320675546303, 0.352174964966252, 0.111072633648291, 0.429328023921698, 0.756713891401887, 0.621890604496002, 0.151855323463678, 0.634948333958164, 0.0887831402942538, 0.789555799448863, 0.110938819358125, 0.610751865897328, 0.0419538165442646, 0.0398274196777493, 0.16824242263101, 0.661169399973005, 0.0578741738572717, 0.10802475665696, 0.224726763321087, 0.0301845124922693, 0.0590566922910511, 0.402787594124675, 0.654286256060004, 0.651807146845385, 0.646300381748006, 0.012573461746797, 0.811203269986436, 0.538836810737848, 0.516833926318213, 0.414405237417668, 0.383745699655265, 0.466365874512121, 0.120775948511437, 0.57918692776002, 0.51449003838934, 0.184409760637209, 0.679985890397802, 0.252696165116504, 0.503243424464017, 0.325467907823622, 0.782800849992782, 0.433969079051167, 0.309705019928515, 0.332402200903744, 0.385025360155851, 0.0146979088895023, 0.213416303973645, 0.249709998490289, 0.85567511129193, 0.061283839866519, 0.217409074539319, 0.321914084721357, 0.41186740109697, 0.124692826531827, 0.400092417839915, 0.577833441784605, 0.190557387657464, 0.774008462205529, 0.584894572151825, 0.215554161928594, 0.20269605377689, 0.363439506152645, 0.787114370847121, 0.565572967752814, 0.35687974630855, 0.669046060414985, 0.441766951000318, 0.722381681669503, 0.936631304211915, 0.0808612513355911, 0.460349786328152, 0.0448870942927897, 0.487027707044035, 0.108414114685729, 0.136052479967475, 0.283654982922599, 0.364086325513199, 0.187032846268266, 0.492983640870079, 0.130885648075491, 0.13617569883354, 0.539112603291869, 0.915092078736052, 0.450533986324444, 0.44665850745514, 0.976848594844341, 0.81193157332018, 0.212711126310751, 0.422790283802897, 0.664532205089927, 0.391243093414232, 0.0250143641605973, 0.835184487048537, 0.136179149150848, 0.14449669001624, 0.894103244645521, 0.599449296947569, 0.262564533157274, 0.136818232946098, 0.77928019175306, 0.964661107864231, 0.142900982871652, 0.00391862401738763, 0.904146103654057, 0.627072693780065, 0.233477899804711, 0.79603413050063, 0.420216082362458, 0.818700607167557, 0.432036695769057, 0.273094281321391, 0.0210732081905007, 0.284269228111953, 0.233133917907253, 0.452670454280451, 0.953142629005015, 0.435082379262894, 0.999271149514243, 0.85069044935517, 0.745869072619826, 0.120188924018294, 0.0264383745379746, 0.603022575844079, 0.365077487658709, 0.228099819738418, 0.537796901306137, 0.306541775818914, 0.874493942130357, 0.342912224587053, 0.552921305177733, 0.49744345061481, 0.619375600712374, 0.697190743405372, 0.0155323974322528, 0.816229894757271, 0.426760418340564, 0.147678921232, 0.126911538653076, 0.835795669583604, 0.695683903060853, 0.992526952875778, 0.761439569061622, 0.539295496419072, 0.985125947045162, 0.666222579777241, 0.0577886488754302, 0.61568214208819, 0.538481465075165, 0.0153307563159615, 0.0178690417669713, 0.0403019413352013, 0.716400602599606, 0.439758623018861, 0.62626748951152, 0.513891724403948, 0.053610784932971, 0.34944949275814, 0.860774309607223, 0.411946158390492, 0.782847341382876, 0.936620099237189, 0.521957836346701, 0.0608568824827671, 0.521432684501633, 0.194629424018785, 0.648225624114275, 0.150464312871918, 0.0412862875964493, 0.0693832824472338, 0.250212842365727, 0.903397712856531, 0.187224556691945, 0.857664785115048, 0.379096639342606, 0.771879525156692, 0.89704358484596, 0.176499836379662, 0.699446859071031, 0.393697289517149, 0.0191986600402743, 0.742280336329713, 0.0229818965308368, 0.89097143150866, 0.0882607789244503, 0.734542956808582, 0.993493812857196, 0.286554357502609, 0.651741255540401, 0.764072358142585, 0.907278445549309, 0.231652815593407, 0.436355289304629, 0.0745267933234572, 0.264589990489185, 0.585478585213423, 0.905375459231436, 0.304067575372756, 0.299749257508665, 0.5023241143208, 0.650710772257298, 0.690253335982561, 0.169479279313236, 0.618792121764272, 0.815320600755513, 0.0450925580225885, 0.191970237297937, 0.00693605281412601, 0.471433915663511, 0.763829422183335, 0.573404447175562, 0.360424726037309, 0.441800808766857, 0.661317366641015, 0.718692220048979, 0.280369525076821, 0.257873008958995, 0.910727800335735, 0.627138777868822, 0.621851807227358, 0.297324220882729, 0.342476218240336, 0.612067013280466, 0.417445985367522, 0.814219921361655, 0.984822983387858, 0.440726315835491, 0.524474919773638, 0.195020135957748, 0.839908178662881, 0.715131175937131, 0.659321028972045, 0.695331540657207, 0.201757018920034, 0.392968945903704, 0.626312558539212, 0.471603531157598, 0.385580196743831, 0.101140995975584, 0.908754400676116, 0.235139077762142, 0.859018611721694, 0.31953398627229, 0.382812165422365, 0.0784714999608696, 0.233843917259946, 0.862700612749904, 0.657812363002449, 0.830557824345306, 0.822702287230641, 0.462563572451472, 0.16839730925858, 0.0700653977692127, 0.145235671661794, 0.208974197506905, 0.133102171123028, 0.147595729446039, 0.377973467810079, 0.351794624933973, 0.156619379995391, 0.709796874318272, 0.11545137828216, 0.697577102808282, 0.745412449585274, 0.337759595829993, 0.863310422748327, 0.386784209404141, 0.0304763289168477, 0.302012245636433, 0.877768979640678, 0.453233288601041, 0.144052239367738, 0.588669790187851, 0.0313734682276845, 0.93279231665656, 0.698926420416683, 0.889840671559796, 0.0237571615725756, 0.639581761322916, 0.744348014239222, 0.0911940420046449, 0.388017242075875, 0.842821442754939, 0.171312117017806, 0.368142791325226, 0.67422836041078, 0.928519306937233, 0.347521256422624, 0.0270392454694957, 0.762900826288387, 0.949952022172511, 0.943705593235791, 0.646099618403241, 0.445993417408317, 0.581701057264581, 0.00896951626054943, 0.226279986090958, 0.647588405525312, 0.827207183232531, 0.00736918020993471, 0.503123984672129, 0.182243980001658, 0.178331731352955, 0.914798323996365, 0.53141515259631, 0.351372905774042, 0.902088741771877, 0.148634069832042, 0.0239711736794561, 0.573332389350981, 0.580123912077397, 0.363417140673846, 0.032121449476108, 0.0759464092552662, 0.189917745534331, 0.811358163366094, 0.0143298185430467, 0.778894656337798, 0.622278351336718, 0.476087902905419, 0.817933312151581, 0.563643920933828, 0.532594347372651, 0.218871075892821, 0.514685429399833, 0.938206668011844, 0.0441015285905451, 0.693191171623766, 0.616833162494004, 0.462613076670095, 0.150132204871625, 0.531597798457369, 0.0900365123525262, 0.398034274112433, 0.701589725911617, 0.939374988432974, 0.487464826786891, 0.434064622269943, 0.32520875451155, 0.794462290825322, 0.75560136558488, 0.751870870823041, 0.633555466076359, 0.157159024849534, 0.689806467620656, 0.404723690357059, 0.0641730800271034, 0.188579654321074, 0.888044669991359, 0.271012621000409, 0.856259665917605, 0.7183888098225, 0.862811149796471, 0.0325229312293231, 0.494465076597407, 0.0387972556054592, 0.991044095950201, 0.541928666876629, 0.782506491290405, 0.229409363586456, 0.65497140889056, 0.634659021394327, 0.325978499138728, 0.301815106533468, 0.794973394367844, 0.314520546235144, 0.570228119613603, 0.658767686691135, 0.117776242317632, 0.0218001180328429, 0.179841683479026, 0.898565171984956, 0.114650929346681, 0.00495814345777035, 0.144443507073447, 0.562547883717343, 0.428854637779295, 0.664980847621337, 0.581862698309124, 0.908705868758261, 0.685318333795294, 0.140287593239918, 0.406389380106702, 0.845643733628094, 0.622236280702055, 0.471408120822161, 0.0208931942470372, 0.624716078862548, 0.0569535903632641, 0.974333015037701, 0.998575335601345, 0.848912509391084, 0.613726488081738, 0.111393547616899, 0.659662895370275, 0.32299708831124, 0.688852351391688, 0.424801806686446, 0.491317426785827, 0.549044752027839, 0.709992456249893, 0.235648683272302, 0.140011162497103, 0.945964619051665, 0.311083472333848, 0.793071992928162, 0.831200960325077, 0.192856245907024, 0.232042279094458, 0.000917298486456275, 0.690897638211027, 0.621474648825824, 0.960666714934632, 0.292254668194801, 0.822349041933194, 0.670681964838877, 0.246562646934763, 0.484690019628033, 0.986233423929662, 0.695113736437634, 0.728531915228814, 0.526909068925306, 0.525852442486212, 0.423893367871642, 0.335489904973656, 0.995627247728407, 0.985253340797499, 0.240783301880583, 0.450798990670592, 0.0494612222537398, 0.0226140473969281, 0.67780352011323, 0.212213923688978, 0.831265797140077, 0.378022553166375, 0.390907981898636, 0.39440935314633, 0.903041349491104, 0.0188181817065924, 0.234965182840824, 0.546790829859674, 0.583126890473068, 0.444764294661582, 0.513866333290935, 0.558746431954205, 0.627252285834402, 0.58408984541893, 0.540356932906434, 0.684927922207862, 0.0182811925187707, 0.964843421708792, 0.284868435468525, 0.938608028227463, 0.802616607630625, 0.995799313066527, 0.518500379053876, 0.10464859334752, 0.991030998062342, 0.317330358317122, 0.614861108828336, 0.453825632110238, 0.0153256801422685, 0.464193169726059, 0.744049973553047, 0.480074883438647, 0.419226342346519, 0.673038359265774, 0.659905357053503, 0.66359581798315, 0.035660985391587, 0.947398165939376, 0.561636554542929, 0.922367845196277, 0.139515401795506, 0.48829959332943, 0.944116011727601, 0.718033086508512, 0.0836298516951501, 0.488948414800689, 0.185351158492267, 0.211180703481659, 0.878565919585526, 0.617826377041638, 0.33196071209386, 0.454417024273425, 0.00382221792824566, 0.858787572477013, 0.336759405676275, 0.100823881337419, 0.474178326549008, 0.599789574276656, 0.199045157525688, 0.0660197103861719, 0.246904556639493, 0.606714085210115, 0.0705452703405172, 0.971427785931155, 0.822556970641017, 0.970208371756598, 0.716878798091784, 0.713594001485035, 0.00676564523018897, 0.055866353912279, 0.341952264076099, 0.6912795023527, 0.034098016563803, 0.847749389940873, 0.377754319692031, 0.422921748599038, 0.324915147852153, 0.327631369000301, 0.368670237949118, 0.966036116238683, 0.867567793931812, 0.876247177366167, 0.487024413887411, 0.398913148790598, 0.728120096493512, 0.125024431152269, 0.659385300474241, 0.274337216746062, 0.0204510558396578, 0.283906876342371, 0.766322669805959, 0.0193104348145425, 0.831395873567089, 0.398025457980111, 0.417655888479203, 0.550837917486206, 0.702506431378424, 0.427801068639383, 0.86251614545472, 0.465509100351483, 0.368411230621859, 0.74950641207397, 0.69199523399584, 0.447477093664929, 0.749112303834409, 0.31889488757588, 0.80313775036484, 0.881201379001141, 0.139681713888422, 0.203259377041832, 0.855569595471025, 0.772459525847808, 0.958454049890861, 0.87457181327045, 0.737379057565704, 0.598655300214887, 0.613269094144925, 0.0066446524579078, 0.434451834531501, 0.521287906449288, 0.896941681392491, 0.4408893447835, 0.45353790698573, 0.83599725109525, 0.166547803906724, 0.145869528641924, 0.298731943359599, 0.736721988767385, 0.71976750693284, 0.288321117637679, 0.247321133268997, 0.416118100984022, 0.743775499984622, 0.21003113524057, 0.795290065929294, 0.128339566523209, 0.729208475910127, 0.931236093863845, 0.347267113858834, 0.323674934916198, 0.00541995419189334, 0.834948576288298, 0.396779139759019, 0.409854518482462, 0.619934174697846, 0.444433351512998, 0.0321722298394889, 0.3974704798311, 0.192036180756986, 0.614968186477199, 0.220879912376404, 0.760391848627478, 0.465508950408548, 0.222679814090952, 0.810082024428993, 0.845483839279041, 0.109894462162629, 0.574257069267333, 0.331721803173423, 0.857940562535077, 0.0423991081770509, 0.879184823017567, 0.0210802431683987, 0.775146696250886, 0.985437799943611, 0.055378551594913, 0.474187796469778, 0.0859081789385527, 0.815720847574994, 0.118043893482536, 0.45022858469747, 0.738034875364974, 0.589378213277087, 0.0884122520219535, 0.78487558895722, 0.506019635125995, 0.694503530627117, 0.283117888262495, 0.686699622543529, 0.312031978275627, 0.522173797246069, 0.539484575390816, 0.160211905371398, 0.963969699107111, 0.0546934655867517, 0.534901855746284, 0.439633695175871, 0.838403965346515, 0.694899442838505, 0.449846395291388, 0.627540776273236, 0.742933761328459, 0.881707831053063, 0.563868915196508, 0.525552605511621, 0.831151692662388, 0.0544589390046895, 0.666572171729058, 0.449990363093093, 0.433930683182552, 0.722376948222518, 0.473962330725044, 0.255476224934682, 0.960011394927278, 0.0522895520552993, 0.63580745505169, 0.664852382149547, 0.941691283835098, 0.589717253809795, 0.129085354274139, 0.540984953986481, 0.712966984137893, 0.508053534897044, 0.92792496480979, 0.525383488740772, 0.145120805129409, 0.568571894895285, 0.710780101828277, 0.485036022495478, 0.270638596732169, 0.735708943102509, 0.599873796803877, 0.812503963243216, 0.680595707846805, 0.952188719063997, 0.577798184938729, 0.950066779041663, 0.969806991750374, 0.70009654504247, 0.0555345499888062, 0.0108837457373738, 0.444467038381845, 0.0642135750968009, 0.889763605548069, 0.923286398174241, 0.726011075312272, 0.0428287631366402, 0.622050745179877, 0.0472844350151718, 0.631144429324195, 0.474510212661698, 0.292617164785042, 0.114854716230184, 0.419389187125489, 0.69663544697687, 0.637733823619783, 0.936399167403579, 0.905558552825823, 0.887996859150007, 0.710645687067881, 0.313472479814664, 0.887091680662706, 0.904292257968336, 0.5844110858161, 0.520035601686686, 0.407476421445608, 0.917648752685636, 0.95769802806899, 0.301626588217914, 0.213700097985566, 0.24380537471734, 0.9527732511051, 0.649852610426024, 0.112305607413873, 0.413368317764252, 0.551665055099875, 0.826136895222589, 0.0887843859381974, 0.672073814319447, 0.915107500506565, 0.253267075167969, 0.72043937491253, 0.859923670999706, 0.118103818269446, 0.71686023613438, 0.150107783731073, 0.0372602227143943, 0.781722167506814, 0.678651015739888, 0.800679271575063, 0.0391206897329539, 0.518377816071734, 0.729962571058422, 0.00621217419393361, 0.651044423924759, 0.873994469642639, 0.331767515977845, 0.937498443527147, 0.490456024883315, 0.752208200516179, 0.154471851885319, 0.500503055751324, 0.0712352071423084, 0.114309768425301, 0.314021012280136, 0.0326338775921613, 0.963038831483573, 0.730518527794629, 0.960837098537013, 0.67900996375829, 0.857125867856666, 0.440895654261112, 0.428971941582859, 0.549614367773756, 0.153366127982736, 0.142892003292218, 0.322108371648937, 0.148272355552763, 0.0504856579937041, 0.248909116256982, 0.0632443726062775, 0.935309870634228, 0.504421031568199, 0.347417760873213, 0.606807957636192, 0.673680182313547, 0.0674395449459553, 0.299415862653404, 0.627867824863642, 0.635108186863363, 0.0672990991733968, 0.873959266580641, 0.844706770032644, 0.314111373620108, 0.939202046720311, 0.846474009566009, 0.252864310983568, 0.496045769192278, 0.598402612609789, 0.330741206649691, 0.757563489954919, 0.91266202647239, 0.960034974385053, 0.400111584924161, 0.344654463697225, 0.0793635484296829, 0.946011599618942, 0.3126923115924, 0.561761994380504, 0.87486335542053, 0.874234500108287, 0.00104550505056977, 0.0750324376858771, 0.311134174233302, 0.933566908352077, 0.0659155843313783, 0.776497649727389, 0.884632158791646, 0.0181933995336294, 0.27066238457337, 0.236168286995962, 0.365644162055105, 0.161440374795347, 0.0176785183139145, 0.686128898058087, 0.698633613763377, 0.484872970497236, 0.516987421084195, 0.416875639231876, 0.115421822294593, 0.165352070238441, 0.698227873072028, 0.636119618779048, 0.214428533334285, 0.699977159267291, 0.954996524145827, 0.807112045353279, 0.000100445467978716, 0.132600004551932, 0.742045800667256, 0.31752431136556, 0.205732540227473, 0.489898150553927, 0.865301775513217, 0.0497566328849643, 0.770952434511855, 0.750385115621611, 0.746349016437307, 0.51926183165051, 0.425770313944668, 0.752731444546953, 0.718250093283132, 0.568825187627226, 0.0060426234267652, 0.688125753309578, 0.861432786565274, 0.695184401469305, 0.813000395894051, 0.700425798771903, 0.151356172282249, 0.341465848730877, 0.855850039515644, 0.528628908796236, 0.292787610786036, 0.624576384201646, 0.0313451357651502, 0.926873431773856, 0.360095396637917, 0.967618949245661, 0.661604127380997, 0.332431876566261, 0.420864413259551, 0.813768558437005, 0.247831605141982, 0.913300792453811, 0.0551167230587453, 0.734224407933652, 0.798658998450264, 0.706044640159234, 0.383956076577306, 0.0244069094769657, 0.781569306040183, 0.740982966031879, 0.152303876820952, 0.243870863458142, 0.811940831132233, 0.835871966090053, 0.283686111215502, 0.718086677836254, 0.593400802928954, 0.0891325336415321, 0.0359444178175181, 0.217448627576232, 0.479849790222943, 0.651884674327448, 0.644836873980239, 0.11269750026986, 0.257218695245683, 0.499611086444929, 0.669005020754412, 0.180745311547071, 0.790595594560727, 0.386579790152609, 0.425397214479744, 0.779146084096283, 0.204563564155251, 0.974167115287855, 0.390916138421744, 0.660567100625485, 0.806502883555368, 0.434864228358492, 0.687703085131943, 0.0884508723393083, 0.352789326338097, 0.0707222337368876, 0.865529701579362, 0.933901228476316, 0.400236577261239, 0.120987423928455, 0.914416919695213, 0.55926447850652, 0.957843634765595, 0.222501351498067, 0.147739016450942, 0.391267829574645, 0.937753702979535, 0.694360339082778, 0.742114585358649, 0.0611800062470138, 0.792978535639122, 0.896184875862673, 0.0106175388209522, 0.555283476598561, 0.151360467076302, 0.117039846954867, 0.761171521618962, 0.943017944460735, 0.204634373309091, 0.175342841539532, 0.178650571266189, 0.16809957427904, 0.417182302102447, 0.69766009808518, 0.486912732245401, 0.863751297118142, 0.715781671693549, 0.117523136548698, 0.480952741811052, 0.585075452690944, 0.0989792256150395, 0.803284467896447, 0.124933501938358, 0.512381877517328, 0.270809234818444, 0.640734900953248, 0.166976110544056, 0.67972770286724, 0.402902650414035, 0.279259070055559, 0.068530393531546, 0.706683478783816, 0.148008976131678, 0.482842483092099, 0.670991631224751, 0.0915371775627136, 0.246239908738062, 0.0742015740834177, 0.739084001397714, 0.393018602626398, 0.183997560525313, 0.725539662409574, 0.924235926475376, 0.935180160216987, 0.0647858395241201, 0.186929616844282, 0.579553705640137, 0.607726008165628, 0.248111336492002, 0.338944152696058, 0.446132970508188), actual = c(0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.591256647137925, 0.326204111566767, 0.892766352975741, 0.0350846399087459, 0.321149893803522, 0.650748055428267, 0.2272086921148, 0.0502748263534158, 0.447762593626976, 0.667494086083025, 0.489607958123088, 0.111098689259961, 0.62382687209174, 0.238195697078481, 0.121085529914126, 0.0458944910205901, 0.0399280749261379, 0.862476144218817, 0.998545325826854, 0.780783607624471, 0.575803821207955, 0.619419697904959, 0.46741694281809, 0.446294002002105, 0.0317435481119901, 0.490406764904037, 0.0281605534255505, 0.119588879635558, 0.703097603516653, 0.989778704708442, 0.837167819263414, 0.963453735923395, 0.575725583359599, 0.138144831405953, 0.718692512484267, 0.485835012514144, 0.42652821703814, 0.560247891116887, 0.135157961864024, 0.673204120015725, 0.314051933586597, 0.4279281408526, 0.866489934269339, 0.940933163976297, 0.0813774152193218, 0.625334279611707, 0.971201266627759, 0.751966357929632, 0.747379406820983, 0.00121092749759555, 0.76390281599015, 0.845746558858082, 0.950609615305439, 0.167154266731814, 0.417177106253803, 0.554655216401443, 0.274337587179616, 0.359084188239649, 0.0130160064436495, 0.136234029661864, 0.104958506766707, 0.3118704801891, 0.659484729170799, 0.819710855837911, 0.282243964960799, 0.195144913159311, 0.25080579565838, 0.326941613573581, 0.574044552864507, 0.379708398366347, 0.0447019620332867, 0.164127101656049, 0.96636820002459, 0.741601914167404, 0.497866345802322, 0.445703998906538, 0.209867830853909, 0.226128303911537, 0.366125211119652, 0.811197054572403, 0.385342448949814, 0.668699691304937, 0.969956618268043, 0.219864362152293, 0.915266514522955, 0.712465251795948, 0.607796362834051, 0.569941272493452, 0.853133446536958, 0.672640362521634, 0.540762842399999, 0.303390123182908, 0.951269274344668, 0.729284045286477, 0.803011703072116, 0.282283414155245, 0.480915358290076, 0.402243335964158, 0.98303296440281, 0.103157486068085, 0.571043980075046, 0.825939713977277, 0.853021113434806, 0.489921527216211, 0.833250411087647, 0.91707967990078, 0.917948785470799, 0.267946750391275, 0.743668987881392, 0.958452999126166, 0.55250216810964, 0.352528124116361, 0.0350695305969566, 0.158709768904373, 0.554899531183764, 0.326680429978296, 0.577248118817806, 0.933112058322877, 0.586651491234079, 0.0666654163505882, 0.728854814777151, 0.270242691738531, 0.554424585076049, 0.227449564496055, 0.744391768239439, 0.952761008869857, 0.277306417003274, 0.68985230429098, 0.915601159445941, 0.634911438450217, 0.70347913983278, 0.691727702040225, 0.783888771431521, 0.952542832586914, 0.0216862100642174, 0.289676323067397, 0.516603894531727, 0.980247030500323, 0.918797910446301, 0.537082558730617, 0.663085125386715, 0.482623501447961, 0.570962696103379, 0.970079392427579, 0.421422490151599, 0.0186867078300565, 0.532068353844807, 0.633746986277401, 0.0854104359168559, 0.472305831499398, 0.032135980669409, 0.142343506915495, 0.137005211785436, 0.943066256586462, 0.698385682422668, 0.506859075743705, 0.481755037093535, 0.940329099772498, 0.529205252882093, 0.228796421084553, 0.519176816567779, 0.144032473675907, 0.820526949828491, 0.954858832759783, 0.307819319423288, 0.396365155931562, 0.415654368931428, 0.237861812580377, 0.900180972414091, 0.477167106233537, 0.649904286488891, 0.281956416089088, 0.170105355791748, 0.405552904354408, 0.298929890617728, 0.812298985663801, 0.442656212486327, 0.819830052088946, 0.176019942155108, 0.876908598700538, 0.185948998201638, 0.954024381237105, 0.107778215315193, 0.219352485612035, 0.976011153543368, 0.234073981177062, 0.543113168329, 0.111485014669597, 0.439024647697806, 0.563480149023235, 0.7895899314899, 0.0710065742023289, 0.496972957858816, 0.70465372921899, 0.254474778892472, 0.0471359300427139, 0.0354649967048317, 0.185076090972871, 0.0362631969619542, 0.621995603200048, 0.617765305796638, 0.873215831117705, 0.0903978974092752, 0.860506198834628, 0.416650771861896, 0.524092745268717, 0.694838099647313, 0.754503253148869, 0.942189367720857, 0.65287013258785, 0.344590530497953, 0.963241818826646, 0.756521965609863, 0.230060323141515, 0.241678427672014, 0.133049056865275, 0.2592923338525, 0.447335192933679, 0.623175790999085, 0.848709228215739, 0.0894798277877271, 0.965740972431377, 0.921196347335353, 0.246416137553751, 0.419478328432888, 0.481892086332664, 0.49485371937044, 0.815828141756356, 0.13338488410227, 0.0870706725399941, 0.165566438576207, 0.622210156172514, 0.163840343244374, 0.502101562917233, 0.800199619028717, 0.20365021773614, 0.650445440551266, 0.296008501201868, 0.410315364366397, 0.236051082145423, 0.495037511456758, 0.675872263498604, 0.0880762410815805, 0.540935522411019, 0.458971281535923, 0.777547352947295, 0.388009114423767, 0.622374495258555, 0.835222809342667, 0.279590093763545, 0.806401582667604, 0.574002797249705, 0.0915869206655771, 0.157133836997673, 0.873620148515329, 0.570102496771142, 0.332884975709021, 0.119878363329917, 0.716836041072384, 0.186503445729613, 0.320441736606881, 0.183512379182503, 0.776541223982349, 0.621663013938814, 0.986618211027235, 0.238694822881371, 0.169585082679987, 0.901923817815259, 0.722705853870139, 0.982445991132408, 0.199846365023404, 0.108220841037109, 0.825345021439716, 0.650393749820068, 0.860971054527909, 0.83525582542643, 0.472718778764829, 0.039253152674064, 0.490966721205041, 0.653339719166979, 0.705617936793715, 0.412121485453099, 0.467930536717176, 0.36586240446195, 0.341671669390053, 0.371928242966533, 0.696898068068549, 0.52157871844247, 0.1636070786044, 0.960180287482217, 0.747557235416025, 0.920795603655279, 0.361373785417527, 0.509439152665436, 0.118662979919463, 0.989059893181548, 0.986568030668423, 0.983177645597607, 0.819784945808351, 0.484988127602264, 0.6059051444754, 0.458435661857948, 0.349151639733464, 0.0509391804225743, 0.662603884469718, 0.370620842091739, 0.312606025487185, 0.381639674305916, 0.707620284287259, 0.121165457414463, 0.470079940278083, 0.575758378719911, 0.875862392829731, 0.731006225571036, 0.0898723159916699, 0.502974041970447, 0.817383743356913, 0.935069274390116, 0.674193833721802, 0.446303270524368, 0.479989046929404, 0.586302610579878, 0.496651900233701, 0.463508512126282, 0.994884085841477, 0.287563737249002, 0.0208986441139132, 0.942700801417232, 0.823627318022773, 0.0677548579405993, 0.196352406637743, 0.496606604661793, 0.892630050424486, 0.970405590953305, 0.910787496017292, 0.39102760469541, 0.24733399390243, 0.951204320183024, 0.808528100606054, 0.0327876207884401, 0.895218886667863, 0.483707746956497, 0.951845767442137, 0.205663952277973, 0.420863838633522, 0.708958081435412, 0.391418447718024, 0.26208313414827, 0.396042713662609, 0.677259715273976, 0.555738195544109, 0.997263032244518, 0.825341386022046, 0.076243193121627, 0.79466865863651, 0.48272041650489, 0.0636057886295021, 0.560738562839106, 0.245020217262208, 0.480003664037213, 0.790900053456426, 0.883530842605978, 0.896172492299229, 0.418795999139547, 0.046963615808636, 0.79415143141523, 0.770457245176658, 0.0380534345749766, 0.784542513079941, 0.456947160651907, 0.463852801593021, 0.172715108375996, 0.267999057658017, 0.294230488827452, 0.281081344000995, 0.380210370291024, 0.11395576945506, 0.199890887830406, 0.372541825519875, 0.778223864268512, 0.922183326911181, 0.736687779426575, 0.803057305980474, 0.94874268793501, 0.614717784337699, 0.307859203778207, 0.0962389172054827, 0.247884093783796, 0.303429067833349, 0.602776092942804, 0.785406861687079, 0.761531345546246, 0.453847279306501, 0.951334583340213, 0.551963907899335, 0.263232811819762, 0.613677040906623, 0.519502230919898, 0.568607616238296, 0.965656912885606, 0.0325513468123972, 0.399768815375865, 0.993833233136684, 0.394379322649911, 0.778264885302633, 0.00492414273321629, 0.165593875339255, 0.113279554061592, 0.624668036121875, 0.37631597253494, 0.336441624909639, 0.536883228225634, 0.0170272686518729, 0.782110030995682, 0.258951822295785, 0.342439766507596, 0.486225565662608, 0.509026717627421, 0.396604670211673, 0.730061423964798, 0.0876891894731671, 0.166320253163576, 0.223188204690814, 0.173794521018863, 0.206065567908809, 0.0569444175343961, 0.627841629786417, 0.971297508804128, 0.459546808153391, 0.130274278111756, 0.0129045567009598, 0.764787466265261, 0.657745017204434, 0.20054898224771, 0.474559593712911, 0.245510864537209, 0.177865032572299, 0.589158956427127, 0.580224056961015, 0.42356768110767, 0.38675757101737, 0.99978340184316, 0.258807301521301, 0.531597953289747, 0.276761556742713, 0.0662481922190636, 0.291941130766645, 0.455782908946276, 0.668276145588607, 0.135974714532495, 0.98897742643021, 0.173423693748191, 0.345485951285809, 0.470491382759064, 0.254182551521808, 0.185804835986346, 0.265318518737331, 0.778458012035117, 0.828118990641087, 0.842222892446443, 0.783912259154022, 0.201868653064594, 0.180491117760539, 0.235624982742593, 0.875249436590821, 0.188796915113926, 0.996055794879794, 0.451125509105623, 0.457582899369299, 0.291554098483175, 0.933733731508255, 0.816137669840828, 0.204570092959329, 0.78202914702706, 0.766226579900831, 0.0882818477693945, 0.762509474530816, 0.940633494639769, 0.205735704628751, 0.737761253025383, 0.458612362388521, 0.435537318931893, 0.944787232903764, 0.822642205748707, 0.495346096809953, 0.957726775668561, 0.0945905752014369, 0.487927163485438, 0.154689213261008, 0.282971400069073, 0.077683188021183, 0.948111255187541, 0.878623214317486, 0.452002489706501, 0.579636223614216, 0.565064722672105, 0.866416291333735, 0.14426285144873, 0.54406528105028, 0.97080060839653, 0.860970726469532, 0.304911061190069, 0.227076155599207, 0.311632888624445, 0.577656320529059, 0.385262012248859, 0.30638122279197, 0.709352149860933, 0.959003722062334, 0.130503048188984, 0.0946495095267892, 0.992228723363951, 0.176581126172096, 0.318061586935073, 0.718725081300363, 0.465298356721178, 0.864916848251596, 0.128069841302931, 0.697179580805823, 0.468353268457577, 0.128578654490411, 0.80612467834726, 0.747521384852007, 0.392282850341871, 0.826353317825124, 0.52558299084194, 0.879310011630878, 0.688368578907102, 0.478716526878998, 0.404406288638711, 0.0838550289627165, 0.481181952171028, 0.176674170186743, 0.865490683121607, 0.75381881184876, 0.431792922550812, 0.124131378950551, 0.0524331186898053, 0.244203791022301, 0.0122950058430433, 0.479303733678535, 0.853708051145077, 0.161980010336265, 0.965138115454465, 0.900331070180982, 0.281563480850309, 0.0586099706124514, 0.185386809287593, 0.75402956828475, 0.850557480473071, 0.113352867774665, 0.10702333971858, 0.671401080209762, 0.737549010664225, 0.122135092038661, 0.985216514673084, 0.486761291511357, 0.630903443554416, 0.228939536958933, 0.0470142802223563, 0.895543209742755, 0.890454064588994, 0.887746141524985, 0.164946453645825, 0.495453575626016, 0.202268670778722, 0.325779292732477, 0.196294930065051, 0.53196138003841, 0.178861049935222, 0.248896622098982, 0.697189801372588, 0.127396784955636, 0.0281851380132139, 0.396858094260097, 0.189585625892505, 0.025191624648869, 0.531370192300528, 0.477129115490243, 0.415987066924572, 0.814298335462809, 0.9409436986316, 0.0959657474886626, 0.968340861145407, 0.418774131452665, 0.465979241766036, 0.786843493813649, 0.283273793291301, 0.499622164992616, 0.0161881961394101, 0.0924761027563363, 0.843749519670382, 0.92891631484963, 0.79970859689638, 0.331132874824107, 0.886611456749961, 0.983524069888517, 0.902170391753316, 0.940060080261901, 0.0580923405941576, 0.49179023434408, 0.719699577661231, 0.179071449674666, 0.694492143811658, 0.152099043596536, 0.651667352300137, 0.260593810584396, 0.23157298984006, 0.303089301334694, 0.24497763812542, 0.24956719391048, 0.625420878175646, 0.807878021849319, 0.605379076674581, 0.223123384872451, 0.385732577415183, 0.864713720045984, 0.4010537751019, 0.264327111653984, 0.702962724259123, 0.114324222784489, 0.0957190671470016, 0.0234839809127152, 0.528245229739696, 0.409790752921253, 0.389533679699525, 0.889448234811425, 0.949473393382505, 0.968964267754927, 0.192388933850452, 0.787303646560758, 0.417262945324183, 0.464571665273979, 0.0232485115993768, 0.583009972702712, 0.889798043994233, 0.265079897362739, 0.390607145382091, 0.898887688759714, 0.623887398978695, 0.0146429010201246, 0.843621789477766, 0.0700451927259564, 0.301938592689112, 0.1178910122253, 0.948338841320947, 0.957300090929493, 0.544524391647428, 0.0664968530181795, 0.0316617342177778, 0.736465134657919, 0.817665008595213, 0.0680850516073406, 0.480674931546673, 0.537959381006658, 0.635218832176179, 0.0688673087861389, 0.993778896518052, 0.144415956689045, 0.416439883643761, 0.198808887973428, 0.723257937235758, 0.55490139266476, 0.224703419255093, 0.131057683378458, 0.726930433651432, 0.44086126703769, 0.156105972127989, 0.536344026681036, 0.421477378113195, 0.45002249837853, 0.643161432817578, 0.16957655036822, 0.356411380460486, 0.662820924073458, 0.972492711618543, 0.910107726696879, 0.729267889168113, 0.942324894014746, 0.552779879188165, 0.376628771889955, 0.243382605491206, 0.74301618267782, 0.0298466451931745, 0.193891348782927, 0.472948517417535, 0.83509001461789, 0.697150439722463, 0.514938174048439, 0.239465765887871, 0.933380373986438, 0.310460799373686, 0.968047056114301, 0.772882203105837, 0.975409301929176, 0.538443713914603, 0.828529850346968, 0.334138948936015, 0.226250922773033, 0.39661339437589, 0.850347149418667, 0.607494153082371, 0.0219042929820716, 0.222026544623077, 0.302992490818724, 0.375714535359293, 0.0452594195958227, 0.342969613149762, 0.80456009414047, 0.712523953989148, 0.240846164757386, 0.238199573010206, 0.315871511818841, 0.32106134458445, 0.400153942406178, 0.315394962672144, 0.627891827840358, 0.481788803590462, 0.537521671503782, 0.789748254232109, 0.410722568398342, 0.206409187987447, 0.0519864491652697, 0.682761821895838, 0.267151308944449, 0.260517580201849, 0.042843570234254, 0.647169938310981, 0.22985806572251, 0.824576372979209, 0.573004649486393, 0.497025694698095, 0.5028330469504, 0.813156148651615, 0.917578403372318, 0.0720861400477588, 0.982027085963637, 0.686093840515241, 0.354159001726657, 0.665299915475771, 0.479839521925896, 0.955272636143491, 0.183526372769848, 0.510271801846102, 0.90231456165202, 0.784911705181003, 0.779419726226479, 0.716869081137702, 0.140727871796116, 0.956232300493866, 0.35800024564378, 0.447855437640101, 0.226738857105374, 0.0558020391035825, 0.784867824986577, 0.571227013133466, 0.428787672659382, 0.555187217658386, 0.00399174448102713, 0.00587974674999714, 0.197470735525712, 0.831207745242864, 0.949062562780455, 0.370907753473148, 0.293869328219444, 0.86145867081359, 0.330858459463343, 0.165532296057791, 0.0290904622524977, 0.155029251938686, 0.688992795301601, 0.0989572266116738, 0.223040458280593, 0.264115998055786, 0.556993548991159, 0.843784620054066, 0.430363669525832, 0.257166270399466, 0.45890700048767, 0.771177474875003, 0.383216532180086, 0.221719307592139, 0.551617555320263, 0.901002826169133, 0.334209247957915, 0.0567070208489895, 0.0955614214763045, 0.296995911048725, 0.345471950946376, 0.375016929581761, 0.373692316934466, 0.613200882915407, 0.327980615431443, 0.692691422067583, 0.0525946861598641, 0.695622126106173, 0.789816506905481, 0.602584445150569, 0.671397771686316, 0.75857523502782, 0.358008403563872, 0.76368657220155, 0.977949556196108, 0.934307020390406, 0.906077938852832, 0.310597694013268, 0.0759650864638388, 0.175270166248083, 0.240403971634805, 0.0361569519154727, 0.960519553162158, 0.559243194991723, 0.677327987039462, 0.567706370726228, 0.603759843623266, 0.572714544599876, 0.675430655945092, 0.903001048136503, 0.521564884576946, 0.330019836081192, 0.0323181324638426, 0.103606479708105, 0.943312462186441, 0.845627270173281, 0.0643866129685193, 0.791183927329257, 0.390208608936518, 0.101982929045334, 0.746084118261933, 0.649959818460047, 0.870372671866789, 0.39579135319218, 0.255742145469412, 0.674855713965371, 0.529150977963582, 0.986356804845855, 0.612625694368035, 0.189985469914973, 0.499623728450388, 0.38363516703248, 0.084720007609576, 0.280378584517166, 0.393768592970446, 0.699655510950834, 0.594750372227281, 0.4483677376993, 0.517176110995933, 0.592281490098685, 0.532071351539344, 0.190327761461958, 0.821227116975933, 0.533822458004579, 0.0458153975196183, 0.794781947042793, 0.419338054722175, 0.0558390237856656, 0.777119857957587, 0.456684379139915, 0.490181231638417, 0.165477917063981, 0.716920592589304, 0.651728349737823, 0.140320345759392, 0.945346814347431, 0.929634161293507, 0.321851203683764, 5.16194850206375e-05, 0.227623250102624, 0.59252592548728, 0.482166097033769, 0.0749537874944508, 0.59378445125185, 0.0377192676533014, 0.360056065488607, 0.0132176270708442, 0.694292373489588, 0.717602196382359, 0.10234078974463, 0.586920861154795, 0.577304544392973, 0.953898081323132, 0.993762864964083, 0.334803476696834, 0.781441720435396, 0.362003949237987, 0.78268202743493, 0.59099190775305, 0.231731637381017, 0.900395627366379, 0.00923645449802279, 0.579875439172611, 0.674962444696575, 0.73878490109928, 0.952475758967921, 0.656473227543756, 0.72394628287293, 0.234640959883109, 0.805340099614114, 0.242378123570234, 0.498689263360575, 0.483202029718086, 0.782759190071374, 0.564468443160877, 0.567489733221009, 0.441532563185319, 0.721050785854459, 0.575700650224462, 0.609487698413432, 0.173222099430859, 0.370268388418481, 0.543437652988359, 0.612620222615078, 0.660879392875358, 0.927373282378539, 0.915727162268013, 0.163928877096623, 0.635068148607388, 0.767977523151785, 0.747037360444665, 0.306970745092258, 0.40618796646595, 0.267688929103315, 0.784778519999236, 0.118247777223587, 0.428644281579182, 0.980551896849647, 0.674550276016816, 0.106404179008678, 0.587630553403869, 0.414260492427275, 0.61760825663805, 0.23010114650242, 0.0122445239685476, 0.0186058878898621, 0.678166161756963, 0.618136958684772, 0.897794136079028, 0.708663452416658, 0.506517344852909, 0.37732324982062, 0.919973484938964, 0.714790990576148, 0.516894444590434, 0.147363303694874, 0.631812195759267, 0.558933045016602, 0.589989089872688, 0.734353753738105, 0.18185627902858, 0.67210347042419, 0.302493954077363, 0.00451325089670718, 0.807614231249318, 0.0502752235624939, 0.0858567215036601, 0.643854567315429, 0.954134980449453, 0.626308551989496, 0.900854874402285, 0.628005218459293, 0.00236336304806173, 0.260316662257537, 0.301016901386902, 0.720746756996959, 0.903239175211638, 0.235780752962455, 0.854024502681568, 0.668464632472023, 0.330228975508362, 0.128792120609432, 0.296747410437092, 0.185834074625745, 0.858615596080199, 0.53847379074432, 0.960480955429375, 0.866507135564461, 0.25679153110832, 0.0853752864059061, 0.727044263388962, 0.469492273405194, 0.244010929949582, 0.300411333097145, 0.793036504648626, 0.269677301403135, 0.600179664324969, 0.591914751799777, 0.474692379822955, 0.434785428224131, 0.562302366131917, 0.0409778857138008, 0.318025080021471, 0.269418478477746, 0.177698257146403, 0.309279215754941, 0.0608454036992043, 0.28057716996409, 0.696675694081932, 0.570682327263057, 0.357432386372238, 0.485008304240182), actual = c(1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"