Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.5083561        0.5072949 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5078255        0.5083561        0.5323445 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.547399709234014, 0.0256733747664839, 0.977613289142027, 0.212193885585293, 0.608147504273802, 0.785450477153063, 0.800692792283371, 0.666667904471979, 0.063515838701278, 0.59149776562117, 0.58828169060871, 0.726738414494321, 0.592759171267971, 0.715165748260915, 0.0168935274705291, 0.136551697272807, 0.416353335371241, 0.544393352931365, 0.351851302431896, 0.187854107934982, 0.848961783573031, 0.110001466469839, 0.643715977901593, 0.0512390832882375, 0.663712670328096, 0.816242905333638, 0.698213646421209, 0.654871265636757, 0.527446372434497, 0.137439785758033, 0.0668482242617756, 0.871663395082578, 0.876701385248452, 0.976388875162229, 0.818852766882628, 0.546608429867774, 0.658312495099381, 0.677098446292803, 0.300609796540812, 0.304013556102291, 0.149578335462138, 0.664091487647966, 0.462867829948664, 0.472037790808827, 0.102496684994549, 0.426418424816802, 0.466925479937345, 0.940591279184446, 0.75743848294951, 0.0409659594297409, 0.844467012677342, 0.375382661353797, 0.346690670354292, 0.687313256785274, 0.880910823121667, 0.226615642430261, 0.951741603203118, 0.337696395348758, 0.511554876575246, 0.990989242913201, 0.66252088942565, 0.89544858946465, 0.431305149802938, 0.18041961081326, 0.175229149637744, 0.686995235737413, 0.964890058385208, 0.661437377799302, 0.71174070588313, 0.946806302992627, 0.718669289723039, 0.86520835920237, 0.871493478771299, 0.673802068224177, 0.326450290624052, 0.243557152338326, 0.307526125106961, 0.415276065003127, 0.0813278099521995, 0.532315003918484, 0.28485158784315, 0.20908685401082, 0.525404229993001, 0.18699838174507, 0.589579975465313, 0.161946475505829, 0.897469211136922, 0.882359619252384, 0.711879087612033, 0.803351839305833, 0.0418444122187793, 0.385192966088653, 0.954081272473559, 0.358458736212924, 0.804195465287194, 0.106556713581085, 0.323791933245957, 0.494198419619352, 0.775448285741732, 0.978086624294519, 0.129947348963469, 0.738147234078497, 0.316248012939468, 0.961157054873183, 0.0094612454995513, 0.882917548296973, 0.0783560401760042, 0.515898767160252, 0.355477316072211, 0.247917168308049, 0.0425115807447582, 0.0792787193786353, 0.393292249646038, 0.567404723493382, 0.080267421901226, 0.352032039081678, 0.684867712669075, 0.0774422183167189, 0.266048279125243, 0.255328907864168, 0.175179450539872, 0.767202039249241, 0.550142274238169, 0.48492016014643, 0.579150406643748, 0.408484651939943, 0.380630591651425, 0.323394510429353, 0.727496699662879, 0.32588218874298, 0.774533654330298, 0.468311094213277, 0.754916070960462, 0.716259209439158, 0.422739709494635, 0.914634306915104, 0.177803472848609, 0.596844170242548, 0.709100894629955, 0.560149512020871, 0.427589246304706, 0.163355392636731, 0.199167556595057, 0.0473749248776585, 0.319622608600184, 0.173752590781078, 0.744708526181057, 0.949212078470737, 0.647959302878007, 0.324954981682822, 0.912307535065338, 0.051176872337237, 0.273427469888702, 0.17666916619055, 0.596698590088636, 0.474132861709222, 0.595918265869841, 0.0857231279369444, 0.608673268463463, 0.113583732163534, 0.359401028836146, 0.0317649475764483, 0.0144664628896862, 0.867263493128121, 0.685666513396427, 0.767605238826945, 0.0116747464053333, 0.833600039593875, 0.881495093926787, 0.0940916230902076, 0.0818311762996018, 0.507729454664513, 0.432277873856947, 0.0137409812305123, 0.338049018988386, 0.549572539981455, 0.978467191802338, 0.534016293706372, 0.873035031370819, 0.716210652142763, 0.787066714605317, 0.108572888886556, 0.856312709627673, 0.787161510204896, 0.122944658156484, 0.236405824776739, 0.522606633836403, 0.798863484757021, 0.546688631176949, 0.621237529907376, 0.601944331079721, 0.906506608007476, 0.237656039651483, 0.642375176772475, 0.516109032789245, 0.358152798842639, 0.687361277639866, 0.0246915321331471, 0.158417148515582, 0.0974687105044723, 0.0223560472950339, 0.626290178624913, 0.75213463511318, 0.242066758684814, 0.0776313573587686, 0.933710779761896, 0.0189781943336129, 0.173079473432153, 0.478627410950139, 0.911188150988892, 0.345125399762765, 0.610184605000541, 0.718971180031076, 0.19406642508693, 0.678663697326556, 0.258796670008451, 0.236856854287907, 0.158492000773549, 0.131089424015954, 0.676062103826553, 0.130640250397846, 0.257748103002086, 0.00992871238850057, 0.933062729658559, 0.504617041209713, 0.819771494949237, 0.632072062464431, 0.449520927621052, 0.759191684657708, 0.0707046794705093, 0.624459179351106, 0.881110665621236, 0.297525041969493, 0.853551016421989, 0.0929339642170817, 0.586940709035844, 0.841449409723282, 0.226833082269877, 0.617628386477008, 0.197735025081784, 0.467398612992838, 0.304379835259169, 0.0420899577438831, 0.252655883785337, 0.646991360699758, 0.13145799911581, 0.496257093269378, 0.744339596480131, 0.542540123919025, 0.678887420101091, 0.0212728446349502, 0.394243391463533, 0.744001126149669, 0.557198893744498, 0.721913506044075, 0.470244541531429, 0.812568913446739, 0.646536055020988, 0.155200434848666, 0.0221056789159775, 0.717321344884112, 0.416779211955145, 0.899293867405504, 0.551531101809815, 0.573731272248551, 0.133340703556314, 0.343247509328648, 0.0982305121142417, 0.141175790689886, 0.709587565390393, 0.400674164295197, 0.913795046741143, 0.945477219298482, 0.634698234964162, 0.553782031172886, 0.34172443440184, 0.134279719321057, 0.713135556317866, 0.633724691811949, 0.251510394737124, 0.520451898686588, 0.299573766998947, 0.276979338377714, 0.749499793630093, 0.971502363216132, 0.300460904836655, 0.251415108330548, 0.364904978312552, 0.674776554573327, 0.375047602923587, 0.225758523214608, 0.763071169378236, 0.552569814957678, 0.344060091534629, 0.997839741175994, 0.648728310363367, 0.701509077567607, 0.878853127360344, 0.691831795033067, 0.671247985679656, 0.427121336106211, 0.138039553072304, 0.319775815587491, 0.377595330122858, 0.580858750734478, 0.72186069842428, 0.120778788812459, 0.188982848078012, 0.495333515806124, 0.14903497020714, 0.286858271807432, 0.729347511660308, 0.806133174104616, 0.769035388249904, 0.0194927705451846, 0.330548274097964, 0.406116048805416, 0.794283572817221, 0.0857651990372688, 0.746018017176539, 0.391741477884352, 0.295764488168061, 0.614762924145907, 0.28130228840746, 0.239233950618654, 0.900237829657272, 0.371237502200529, 0.566803855355829, 0.687546433880925, 0.209213858470321, 0.576620061416179, 0.541424291906878, 0.237040889915079, 0.0967728353571147, 0.75991661963053, 0.467440742766485, 0.964801665162668, 0.491909960983321, 0.505908376537263, 0.537386929616332, 0.871829624753445, 0.331047449726611, 0.961704927962273, 0.810736124403775, 0.557700975332409, 0.197065856773406, 0.776966867735609, 0.281894308282062, 0.276928584091365, 0.568713926244527, 0.357477247016504, 0.0691393294837326, 0.796507396036759, 0.985013299388811, 0.160323971882463, 0.0511148495133966, 0.273189696017653, 0.369220565073192, 0.065696301870048, 0.669123088009655, 0.806124535389245, 0.83556147175841, 0.472903800429776, 0.742624184116721, 0.223901459481567, 0.245684887981042, 0.425321479095146, 0.841141310287639, 0.215950564946979, 0.351156596792862, 0.0642576559912413, 0.893607240635902, 0.245697170961648, 0.0118107376620173, 0.347537818364799, 0.428707476938143, 0.987707570195198, 0.215984102804214, 0.356912060873583, 0.107904583448544, 0.708204541821033, 0.0705784682650119, 0.269520078552887, 0.971220805542544, 0.162235799944028, 0.723933044355363, 0.632869702763855, 0.917245717719197, 0.944808674510568, 0.644300990737975, 0.208057772601023, 0.695125331869349, 0.0079754926264286, 0.107026204932481, 0.2761078432668, 0.550271356478333, 0.986110867699608, 0.902133998693898, 0.0160222216509283, 0.911884495522827, 0.0484330067411065, 0.212604975793511, 0.472165431827307, 0.46074829204008, 0.0574762297328562, 0.90730817290023, 0.667985375737771, 0.959440834121779, 0.27491887146607, 0.915920910192654, 0.0345279241446406, 0.591238899854943, 0.422776280902326, 0.443971663014963, 0.221874517621472, 0.239628534764051, 0.915991965914145, 0.405815750360489, 0.642959747463465, 0.305273949867114, 0.386708637233824, 0.302694195182994, 0.606785465497524, 0.552109030075371, 0.574141816934571, 0.740050010150298, 0.3110672312323, 0.442157219629735, 0.147444622823969, 0.0101552568376064, 0.702569204615429, 0.259516293881461, 0.00509851146489382, 0.353239516727626, 0.74718301743269, 0.27548343106173, 0.00986770074814558, 0.0986479539424181, 0.307189377956092, 0.39044323307462, 0.0528546862769872, 0.0668214720208198, 0.719617707189173, 0.970037824008614, 0.343573287362233, 0.538757266011089, 0.629194404929876, 0.532516065286472, 0.983992245979607, 0.517027623485774, 0.995581542141736, 0.400887640193105, 0.807993796188384, 0.208499796921387, 0.978536210255697, 0.597267139703035, 0.812999538145959, 0.484401039546356, 0.439509122399613, 0.84583132318221, 0.91288323351182, 0.983216788619757, 0.545299255289137, 0.777650521602482, 0.931910179089755, 0.852595433825627, 0.716259805252776, 0.178107810672373, 0.697851434117183, 0.0200656070373952, 0.712526504648849, 0.539523273473606, 0.760186285013333, 0.252696768613532, 0.952877928270027, 0.215314901666716, 0.313209482468665, 0.428538406034932, 0.636247974354774, 0.634973889216781, 0.530907132197171, 0.453690947499126, 0.779102305416018, 0.314355248818174, 0.654533650493249, 0.657420984003693, 0.386912920279428, 0.422279971186072, 0.766985390800983, 0.371263396926224, 0.566194441635162, 0.585440993541852, 0.653073131805286, 0.317826708313078, 0.369956770213321, 0.131127477623522, 0.211488565197214, 0.207885026931763, 0.0927566550672054, 0.341973403468728, 0.189544879132882, 0.892728528939188, 0.622028900077567, 0.880077833076939, 0.769730015890673, 0.97010134300217, 0.107240113196895, 0.869209452532232, 0.871638148790225, 0.979941039113328, 0.270854511763901, 0.98897625785321, 0.578863160684705, 0.76540171331726, 0.491554985754192, 0.807903267676011, 0.934888860676438, 0.100779756205156, 0.358212813036516, 0.908648311626166, 0.247253183973953, 0.473114105407149, 0.328324017114937, 0.00735756661742926, 0.897865805774927, 0.139051876030862, 0.837142186937854, 0.00875973491929471, 0.975532231852412, 0.424741487717256, 0.10528189712204, 0.346463058143854, 0.789696882246062, 0.294916358310729, 0.993239267729223, 0.493645017500967, 0.605739378836006, 0.823503630235791, 0.907983259996399, 0.821059030480683, 0.236807137029245, 0.645052515668795, 0.565336202736944, 0.661613087169826, 0.211602166993544, 0.457971035735682, 0.561326358001679, 0.357109412318096, 0.621971223736182, 0.755797798279673, 0.173475365852937, 0.765396901406348, 0.575442468049005, 0.265815015882254, 0.54251927905716, 0.0209239022806287, 0.85791456955485, 0.0457560168579221, 0.517024479573593, 0.135998539393768, 0.546934198122472, 0.0869953546207398, 0.807431423803791, 0.440592986298725, 0.996597954072058, 0.399700526148081, 0.0710090836510062, 0.212377168238163, 0.914605092024431, 0.0580294637475163, 0.241959895007312, 0.0856848680414259, 0.866217421833426, 0.841461139032617, 0.507284588180482, 0.723465928109363, 0.677405390655622, 0.489614566555247, 0.616897782078013, 0.658148513408378, 0.431956222979352, 0.88206254132092, 0.292103259358555, 0.327688875840977, 0.157882418483496, 0.345510208979249, 0.0906063073780388, 0.00893366825766861, 0.0362484513316303, 0.545202244073153, 0.0809396079275757, 0.316979743773118, 0.35099087189883, 0.963977312203497, 0.657994121778756, 0.549490293022245, 0.48276923969388, 0.473390273982659, 0.940291566075757, 0.751059107016772, 0.917065370129421, 0.906897184439003, 0.447136368602514, 0.601643257541582, 0.876367755932733, 0.822900339728221, 0.990438515553251, 0.00317806703969836, 0.0195088798645884, 0.473525642650202, 0.531170062953606, 0.405850573908538, 0.237679652636871, 0.359541113488376, 0.352257484337315, 0.525736439274624, 0.904619420180097, 0.313372459961101, 0.940464527579024, 0.597288131713867, 0.324534924700856, 0.217762325424701, 0.600425442913547, 0.572128630941734, 0.0686923319008201, 0.797733049374074, 0.168119886890054, 0.670595858013257, 0.710632933536544, 0.159439121140167, 0.360100286779925, 0.305902529507875, 0.0995114580728114, 0.552269569365308, 0.384912147419527, 0.911116851726547, 0.530747780110687, 0.00745802070014179, 0.823721983470023, 0.504543291404843, 0.238716860301793, 0.788208013633266, 0.535693309269845, 0.536866869078949, 0.502445353893563, 0.943816408747807, 0.000494318082928658, 0.720229470636696, 0.106334384065121, 0.706819919869304, 0.797749922610819, 0.604607057757676, 0.268152801785618, 0.708267730195075, 0.630330480402336, 0.617295781848952, 0.693610981572419, 0.765520060202107, 0.707530490122736, 0.965385161340237, 0.549779298249632, 0.424239843618125, 0.671004541451111, 0.879725153557956, 0.475597399054095, 0.870855813380331, 0.165723872371018, 0.762403196189553, 0.858956909971312, 0.327710114652291, 0.265894210897386, 0.107689075171947, 0.895345365395769, 0.451718279160559, 0.50927749206312, 0.720325716771185, 0.226951005868614, 0.998162114527076, 0.74114506249316, 0.010299215791747, 0.00271103996783495, 0.306413002777845, 0.293132101884112, 0.135116682155058, 0.635129507165402, 0.925651190336794, 0.0191658136900514, 0.304437343263999, 0.0208217205945402, 0.387079706415534, 0.5972491633147, 0.0448914722073823, 0.962996711954474, 0.247171817813069, 0.914407249540091, 0.574354161741212, 0.897407398791984, 0.16153577901423, 0.985482326475903, 0.0876739288214594, 0.0372578697279096, 0.430306281661615, 0.229197939392179, 0.712696232134476, 0.991239794064313, 0.811294814106077, 0.573248844360933, 0.257345346733928, 0.215073312399909, 0.876006554812193, 0.647219425067306, 0.0714377043768764, 0.15347698633559, 0.800968804163858, 0.48455709614791, 0.321376327425241, 0.329339686548337, 0.536104705184698, 0.65140601573512, 0.323904767865315, 0.722240848699585, 0.981029413640499, 0.851963215041906, 0.111217632424086, 0.68234768207185, 0.993935750797391, 0.510869931895286, 0.62444302928634, 0.594857311109081, 0.266362631460652, 0.0429058091249317, 0.781161754159257, 0.160860911943018, 0.847664281493053, 0.865309947403148, 0.52054107002914, 0.539299214724451, 0.282724040094763, 0.642985561164096, 0.689727963181213, 0.784451768733561, 0.386407230049372, 0.497041549300775, 0.898461425909773, 0.894098123768345, 0.123028201516718, 0.803380197845399, 0.8221054750029, 0.952414748491719, 0.308042316697538, 0.760523153003305, 0.242730156052858, 0.256605155067518, 0.00636731390841305, 0.336852236418054, 0.778103337623179, 0.915260584326461, 0.31362007278949, 0.264133450808004, 0.142077620374039, 0.983682097867131, 0.432046607602388, 0.432209665188566, 0.326804413693026, 0.221283718012273, 0.0978655535727739, 0.189748122589663, 0.682273138081655, 0.410766882821918, 0.490228927694261, 0.600153596373275, 0.729108425555751, 0.203109450172633, 0.54258149350062, 0.27539587719366, 0.383532298495993, 0.335240963613614, 0.333102639066055, 0.901150786783546, 0.527460598153993, 0.510019160341471, 0.637943419860676, 0.156294198473915, 0.803296212572604, 0.936831830302253, 0.76764937536791, 0.960905249696225, 0.355684162117541, 0.336378659820184, 0.270519856829196, 0.0191477914340794, 0.051610384369269, 0.876014478737488, 0.129792542895302, 0.957644425099716, 0.679485300090164, 0.521832497557625, 0.57525756303221, 0.700546078383923, 0.172600716585293, 0.859701982233673, 0.416211977833882, 0.624916349770501, 0.653429046506062, 0.252596035366878, 0.369475292507559, 0.472064435947686, 0.0219414355233312, 0.575997454114258, 0.295930876163766, 0.830818883609027, 0.185725636547431, 0.725856482051313, 0.722163375001401, 0.781895926687866, 0.702247543260455, 0.539185744710267, 0.388195626437664, 0.374154450604692, 0.87714690505527, 0.722675262019038, 0.075267108855769, 0.173019214998931, 0.363495729630813, 0.254930956754833, 0.136816581245512, 0.937364767072722, 0.882799446117133, 0.944936460116878, 0.0173599126283079, 0.934280034620315, 0.701986562926322, 0.039560052100569, 0.396440606098622, 0.987206784775481, 0.166111857397482, 0.541735697071999, 0.479679787997156, 0.170344991609454, 0.132085371064022, 0.489584245719016, 0.19563875021413, 0.320077407872304, 0.0679714013822377, 0.455234715016559, 0.175875124987215, 0.335129226092249, 0.45470806886442, 0.387486380059272, 0.940425179898739, 0.935613323235884, 0.736296509858221, 0.66450033779256, 0.091810328187421, 0.526452426798642, 0.194055726518854, 0.537211864488199, 0.639117378508672, 0.983149539213628, 0.29118841397576, 0.988480317639187, 0.0857345936819911, 0.435320045100525, 0.793797261780128, 0.465633365791291, 0.754003070760518, 0.159693227149546, 0.748614156851545, 0.73777493624948, 0.786012381780893, 0.669231767766178, 0.206610147608444, 0.0351307431701571, 0.173930766060948, 0.363157768500969, 0.957598634529859, 0.796572190243751, 0.220264012692496, 0.324275962775573, 0.877179229166359, 0.0296066692098975, 0.89636120153591, 0.72297239000909, 0.179450183175504, 0.275937974452972, 0.170401231851429, 0.911075740586966, 0.127627920126542, 0.412782049272209, 0.399284907151014, 0.277858094777912, 0.613155162893236, 0.347620981745422, 0.291970629710704, 0.403319825185463, 0.371360996039584, 0.612694388488308, 0.611243134131655, 0.387447311077267, 0.483532845973969, 0.251041346229613, 0.888353240676224, 0.577700664289296, 0.952994757797569, 0.323577478993684, 0.00845641759224236, 0.954173600068316, 0.289220331702381, 0.663948602275923, 0.521096292184666, 0.39982266863808, 0.432835960527882, 0.040402072481811, 0.885283087613061, 0.594658264191821, 0.851350058801472, 0.0743487405125052, 0.756165334256366, 0.0136142577975988, 0.814951735083014, 0.0408921262715012, 0.498821990564466, 0.662679952569306, 0.885876955231652, 0.404947356320918, 0.86788055463694, 0.82497020252049, 0.938101199455559, 0.131048109615222, 0.230621369555593, 0.304019452305511, 0.75431480910629, 0.882680251030251, 0.965119832893834, 0.0869754757732153, 0.500500206602737, 0.121957545401528, 0.230821106815711, 0.201383697567508, 0.153039961121976, 0.970885241637006, 0.571324696298689, 0.302812191192061, 0.049588963855058, 0.669731233036146, 0.825092856539413, 0.00550408684648573, 0.889693812001497, 0.0652765783015639, 0.674144804710522, 0.512875350657851, 0.871697996975854, 0.45688786287792, 0.115819765720516, 0.642864717170596, 0.678232501493767, 0.73526852321811, 0.523169573629275, 0.480889534577727, 0.210012919735163, 0.738071660278365, 0.0732473158277571, 0.393096251180395, 0.110389478737488, 0.323030017316341, 0.906074394937605, 0.352066878695041, 0.196879440220073, 0.156552126863971, 0.248396061826497, 0.508499426534399, 0.0981673994101584, 0.318091925466433, 0.0908402579370886, 0.478428091621026, 0.873359786346555, 0.665078976890072, 0.278980395756662, 0.601306139491498, 0.141793821705505, 0.552111372118816, 0.938677413854748, 0.786973235663027, 0.0137277410831302, 0.462476113345474, 0.381188259925693, 0.0355866295285523, 0.904446806292981, 0.918098867870867, 0.967620303854346, 0.0582123966887593, 0.637074297759682, 0.559339998988435, 0.0385144725441933, 0.952105820644647, 0.488689340651035, 0.0325619764626026, 0.55703677656129, 0.0614966158755124, 0.25105037051253), actual = c(0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.0216062045656145, 0.114783828146756, 0.538610412972048, 0.339275982696563, 0.211275150999427, 0.0641440597828478, 0.0645411349833012, 0.560851954855025, 0.787626720033586, 0.287992129800841, 0.24813445773907, 0.462339745135978, 0.818562876200303, 0.331874069059268, 0.09167938423343, 0.258667611749843, 0.188211191212758, 0.520867804763839, 0.0132264553103596, 0.0738362479023635, 0.0761167069431394, 0.726926756091416, 0.461771216476336, 0.273495058063418, 0.215889346785843, 0.819389344193041, 0.263533939840272, 0.152006961405277, 0.285957219311967, 0.889072977006435, 0.627989757107571, 0.405896978685632, 0.0520052560605109, 0.350661601405591, 0.0415270566008985, 0.0921026328578591, 0.965272353496403, 0.570415047230199, 0.744016205659136, 0.48217155598104, 0.0887387010734528, 0.24224710627459, 0.0620793993584812, 0.430470636580139, 0.890562194632366, 0.0923154267948121, 0.65435151825659, 0.641957148443907, 0.772818228462711, 0.970079438760877, 0.569542282028124, 0.422542712185532, 0.728971559787169, 0.418731672456488, 0.00133962649852037, 0.57920581754297, 0.920819524908438, 0.685004296712577, 0.597633508732542, 0.880293549969792, 0.938983282074332, 0.191533704521134, 0.735161289107054, 0.474842615891248, 0.42331449338235, 0.300434512086213, 0.775931510142982, 0.31420562393032, 0.724112084601074, 0.889280883129686, 0.175515717128292, 0.645278797484934, 0.139738297788426, 0.909339466132224, 0.406578127294779, 0.96437046257779, 0.618903716327623, 0.915408236673102, 0.126629748847336, 0.142715682741255, 0.02071233978495, 0.915625647641718, 0.378653870662674, 0.295423981035128, 0.237598406383768, 0.421947458060458, 0.484114699298516, 0.019317613216117, 0.958513122983277, 0.0391282935161144, 0.426349811023101, 0.706967733800411, 0.606373185524717, 0.0565757965669036, 0.831327219959348, 0.657704382669181, 0.748735463246703, 0.986803250620142, 0.514158512232825, 0.81252044509165, 0.663028240436688, 0.809392132796347, 0.738797828089446, 0.125226056668907, 0.00557289039716125, 0.868193333037198, 0.49197926511988, 0.781823315890506, 0.21397699508816, 0.87744539603591, 0.264265243429691, 0.645949846599251, 0.355159040773287, 0.668956963345408, 0.824782385956496, 0.298325679730624, 0.66455378732644, 0.350350182037801, 0.00154677778482437, 0.653125856304541, 0.875899235717952, 0.275848201010376, 0.909800782799721, 0.434339832514524, 0.706192007288337, 0.829214373836294, 0.361998729174957, 0.329451486235484, 0.939881007419899, 0.953882714733481, 0.0966603592969477, 0.0118517803493887, 0.991043954622, 0.533406043192372, 0.550476666307077, 0.422559269936755, 0.732674370054156, 0.829230888979509, 0.376809207256883, 0.829243834130466, 0.247371442150325, 0.730758468620479, 0.87957354192622, 0.497097776038572, 0.532111351611093, 0.270039482042193, 0.166493223281577, 0.863849827554077, 0.718222204130143, 0.0496917618438601, 0.880730914184824, 0.487138805212453, 0.412048232741654, 0.84138128045015, 0.105702651664615, 0.123932740651071, 0.342176241334528, 0.776898941257969, 0.926656112773344, 0.287365427939221, 0.364366677356884, 0.670679692644626, 0.992172072641551, 0.913222840288654, 0.589083739090711, 0.116898111766204, 0.428649618057534, 0.713195618474856, 0.284932583803311, 0.917165811173618, 0.193929996108636, 0.389173245523125, 0.570460183545947, 0.47989108460024, 0.387886664131656, 0.343174065696076, 0.979347687913105, 0.0923972562886775, 0.493954217759892, 0.946090666344389, 0.880976520711556, 0.173107881098986, 0.427595834480599, 0.186836503446102, 0.88205751683563, 0.211008690763265, 0.292306845774874, 0.365753397345543, 0.816760736284778, 0.366768604144454, 0.67453837278299, 0.861336646834388, 0.949140437645838, 0.46902606729418, 0.982829706277698, 0.292428771499544, 0.915147200459614, 0.283944063819945, 0.691115400521085, 0.47784163034521, 0.116569471545517, 0.146300175925717, 0.852468512253836, 0.69887138181366, 0.00824495614506304, 0.522663825424388, 0.340261831646785, 0.62566827936098, 0.301032925955951, 0.870437903795391, 0.433399299159646, 0.222025096882135, 0.0947161670774221, 0.729051542468369, 0.488981894683093, 0.135702748782933, 0.216006629634649, 0.342202187748626, 0.415770991006866, 0.506494799628854, 0.87979040411301, 0.0109298252500594, 0.113773769931868, 0.827112187631428, 0.90540294861421, 0.350178055465221, 0.422308770241216, 0.683741098968312, 0.881564900046214, 0.219777064863592, 0.0177601333707571, 0.466872670454904, 0.401810476090759, 0.235524989431724, 0.0686784607823938, 0.709286355879158, 0.573001286014915, 0.429582074750215, 0.560291510540992, 0.940331038786098, 0.763720270246267, 0.738755778875202, 0.150571421720088, 0.264064115239307, 0.961323502007872, 0.0172300147823989, 0.752463272772729, 0.0592709360644221, 0.321013788925484, 0.533438992220908, 0.313475631177425, 0.908923289971426, 0.490333107067272, 0.321898626629263, 0.808930706465617, 0.581535799428821, 0.169380974955857, 0.874165322398767, 0.904457549098879, 0.531093715224415, 0.963143043685704, 0.0365353918168694, 0.812762216199189, 0.925902117742226, 0.584128861548379, 0.359431125689298, 0.916535576339811, 0.806358713423833, 0.394152158172801, 0.979600196005777, 0.303948814747855, 0.52825720468536, 0.699977615848184, 0.799664317397401, 0.00351693108677864, 0.734867274528369, 0.502860335167497, 0.972959287464619, 0.674577413126826, 0.953132421709597, 0.884198266547173, 0.018434235593304, 0.398365292232484, 0.943751259008422, 0.237081030150875, 0.170949071412906, 0.0622452604584396, 0.736629977822304, 0.256997488671914, 0.564862873405218, 0.78120352467522, 0.516710513271391, 0.159647524822503, 0.335729496087879, 0.766386083094403, 0.532958219759166, 0.0111323897726834, 0.502170756226406, 0.3204488188494, 0.663334893295541, 0.31843413063325, 0.87192293908447, 0.909771546721458, 0.0300009143538773, 0.57314671529457, 0.25027181697078, 0.241328845499083, 0.116302302805707, 0.245895229512826, 0.335105160716921, 0.29545294935815, 0.794574898667634, 0.559142030077055, 0.735603510169312, 0.238544343970716, 0.159395905211568, 0.457448883214965, 0.42023514653556, 0.434184420621023, 0.213947089388967, 0.388735804939643, 0.0882794633507729, 0.167579773347825, 0.392966318177059, 0.353596518514678, 0.782636933261529, 0.356557505670935, 0.22239472810179, 0.75421045999974, 0.568268654868007, 0.592120085144415, 0.719790867529809, 0.501664872514084, 0.260683956556022, 0.282890384783968, 0.138215451734141, 0.502208225894719, 0.429612318985164, 0.326832981547341, 0.914343212032691, 0.225563166895881, 0.5171593001578, 0.20909824105911, 0.502196802059188, 0.83950703102164, 0.693113235291094, 0.8539882467594, 0.816168142016977, 0.0661224867217243, 0.874351640231907, 0.427049002144486, 0.829661585623398, 0.371708187507465, 0.519409896340221, 0.204178899526596, 0.616932311095297, 0.305885966867208, 0.699688671389595, 0.439378224080428, 0.866702071158215, 0.0970492351334542, 0.428790643578395, 0.875169454608113, 0.658043460687622, 0.211055263411254, 0.224618736421689, 0.749020786723122, 0.899674442829564, 0.410045258468017, 0.903317589079961, 0.10249104257673, 0.231145932804793, 0.65024346485734, 0.389655560255051, 0.314847494242713, 0.65923313004896, 0.542528455145657, 0.152283864095807, 0.19523948431015, 0.584341899259016, 0.747162381652743, 0.613793555879965, 0.446946792537346, 0.612483870238066, 0.592605358920991, 0.619571209186688, 0.273190509527922, 0.0893768607638776, 0.468972988892347, 0.308011706452817, 0.385542194126174, 0.173832284053788, 0.586316457251087, 0.819157703081146, 0.118482836522162, 0.0880186471622437, 0.548129294300452, 0.148118117125705, 0.944596321322024, 0.764399672858417, 0.894220311194658, 0.566435316344723, 0.728007236495614, 0.987177729373798, 0.119299457641318, 0.256413910537958, 0.941949672996998, 0.202602078905329, 0.712666067294776, 0.208133451873437, 0.823931176913902, 0.0394098544493318, 0.338823162252083, 0.942550006788224, 0.595599948195741, 0.894826677395031, 0.741928420029581, 0.24828879837878, 0.229836756363511, 0.762649976881221, 0.449937304481864, 0.263986286008731, 0.859965643379837, 0.216347401496023, 0.472451368812472, 0.640840077307075, 0.283431368181482, 0.372985007474199, 0.275787519291043, 0.156752835260704, 0.800775730516762, 0.481325627770275, 0.091139605268836, 0.450624275952578, 0.324651591014117, 0.252410966204479, 0.151699293404818, 0.529266177909449, 0.79709770064801, 0.873854928417131, 0.720479893265292, 0.176463010022417, 0.657196638872847, 0.96359665109776, 0.0488509563729167, 0.0625904547050595, 0.712337188655511, 0.144541091751307, 0.678532699355856, 0.728584187105298, 0.0437042478006333, 0.687359048053622, 0.462683151010424, 0.785325973061845, 0.882122846087441, 0.611004454782233, 0.386448706267402, 0.444543940247968, 0.860840051900595, 0.758930209791288, 0.110286441864446, 0.569130076793954, 0.46506836777553, 0.385964675340801, 0.724090872332454, 0.606637033401057, 0.0610278132371604, 0.419010766549036, 0.0513463381212205, 0.989770840154961, 0.887352991150692, 0.127941284095868, 0.404287869809195, 0.894560266751796, 0.306537480326369, 0.764636336825788, 0.673689127899706, 0.672479692613706, 0.34128844877705, 0.533992062788457, 0.904932482400909, 0.58908785157837, 0.997589299455285, 0.0144552243873477, 0.587918460136279, 0.436360054183751, 0.93575941468589, 0.648307281080633, 0.28506402974017, 0.507286893436685, 0.0718749586958438, 0.226087143877521, 0.0356902012135834, 0.610996026545763, 0.886947605758905, 0.70774308219552, 0.152413289761171, 0.0462906041648239, 0.599872739054263, 0.55671556503512, 0.428538186941296, 0.545527860056609, 0.387238539056852, 0.0665412677917629, 0.323180551407859, 0.483109792228788, 0.862758278381079, 0.0629955963231623, 0.996707441750914, 0.690572965890169, 0.573527332860976, 0.454481419408694, 0.346529317088425, 0.343581447843462, 0.958390344167128, 0.391161710955203, 0.156071566278115, 0.622985194902867, 0.131849441211671, 0.592974418541417, 0.230867487611249, 0.851696014637128, 0.654328598175198, 0.291493785800412, 0.0401336452923715, 0.34326793346554, 0.818705764133483, 0.975283714244142, 0.761355884140357, 0.107692078221589, 0.969720199238509, 0.266799622215331, 0.191519399406388, 0.709970930358395, 0.62215154664591, 0.235722229350358, 0.232325894990936, 0.471980038331822, 0.696157810743898, 0.431209002854303, 0.492521435720846, 0.46205343096517, 0.473302385536954, 0.0994289757218212, 0.757497001206502, 0.970732948975638, 0.0945098418742418, 0.985068789217621, 0.800394906895235, 0.37582098855637, 0.824891098542139, 0.0571386504452676, 0.333140058675781, 0.0273264581337571, 0.35530028748326, 0.78304956573993, 0.803635162301362, 0.747610367601737, 0.323244308354333, 0.966185736702755, 0.925767997745425, 0.493603553622961, 0.341470630373806, 0.27921693935059, 0.477022919803858, 0.936640561558306, 0.245073647238314, 0.684736214810982, 0.849409237038344, 0.63280442985706, 0.806962999282405, 0.900990260066465, 0.107975416118279, 0.156148293754086, 0.529175511328503, 0.5714886484202, 0.392715191934258, 0.992590758251026, 0.774354294873774, 0.873764891177416, 0.724841725546867, 0.51082037598826, 0.953810822684318, 0.687405043514445, 0.757604864891618, 0.9101049057208, 0.861036575865, 0.630691497353837, 0.833903339458629, 0.559210209874436, 0.862666795495898, 0.25426624994725, 0.274993092287332, 0.880749094765633, 0.868265876080841, 0.210769539931789, 0.824231598991901, 0.938557093497366, 0.406871809158474, 0.505282103084028, 0.51228278898634, 0.656029141042382, 0.84848128631711, 0.0453230494167656, 0.106594565790147, 0.000496027059853077, 0.205309146316722, 0.475011166650802, 0.897457280894741, 0.877981691621244, 0.837715534260496, 0.839410217246041, 0.623381499666721, 0.324194545391947, 0.238106912467629, 0.545277152443305, 0.563578561646864, 0.393129484262317, 0.850277564954013, 0.250909782713279, 0.0145513685420156, 0.138388807186857, 0.276649537496269, 0.472063269698992, 0.928849073825404, 0.657174845226109, 0.171360245672986, 0.59557936177589, 0.838853976223618, 0.949223572155461, 0.363098811125383, 0.667114698793739, 0.727840188657865, 0.0601739720441401, 0.709129743743688, 0.283795781200752, 0.2467591536697, 0.77155611035414, 0.0468599279411137, 0.287795368116349, 0.442202802281827, 0.141878385329619, 0.844129008008167, 0.935379900503904, 0.285407301504165, 0.337202889611945, 0.44440897135064, 0.259857624769211, 0.0129332961514592, 0.936437513679266, 0.593589840224013, 0.964542584028095, 0.764611095190048, 0.374597274698317, 0.766791288740933, 0.897963091731071, 0.66662220726721, 0.681298996554688, 0.214356727898121, 0.776377528207377, 0.143684047972783, 0.110353485913947, 0.251924022333696, 0.15925028314814, 0.705768706742674, 0.326286803232506, 0.913663690211251, 0.610759159550071, 0.224891245132312, 0.070519516710192, 0.253448927775025, 0.0622803731821477, 0.0591815400402993, 0.592300792923197, 0.321240734308958, 0.370086137205362, 0.019150429405272, 0.0904300229158252, 0.935494086705148, 0.984008534112945, 0.022752158343792, 0.271888015791774, 0.49092248827219, 0.405587704153731, 0.506153471767902, 0.373655041214079, 0.240500528365374, 0.376468567410484, 0.775991244474426, 0.651464735623449, 0.497804349055514, 0.53844449413009, 0.305695103248581, 0.205430914647877, 0.0167817429173738, 0.99886471265927, 0.616844401694834, 0.199646713677794, 0.13016469613649, 0.212848875438794, 0.608036153251305, 0.803731155814603, 0.307810774538666, 0.64778853720054, 0.135314106941223, 0.414455406367779, 0.167363595450297, 0.00382698187604547, 0.00492846057750285, 0.0446060688700527, 0.577599602518603, 0.418094928143546, 0.165038906736299, 0.474487372906879, 0.615079001756385, 0.012495705159381, 0.969360520364717, 0.107372984755784, 0.0813371930271387, 0.471223424654454, 0.575950467959046, 0.354977126000449, 0.359155248617753, 0.974909409182146, 0.39398509520106, 0.124438353814185, 0.904454130446538, 0.797504930524155, 0.785457378253341, 0.767830503406003, 0.0688012249302119, 0.41991002112627, 0.296189316781238, 0.874429464805871, 0.916738390689716, 0.397182274144143, 0.898265779484063, 0.195769815240055, 0.925524843856692, 0.680951623013243, 0.376930484781042, 0.134239616803825, 0.629846374737099, 0.68650787579827, 0.0588738196529448, 0.546181900193915, 0.134067832957953, 0.477392199216411, 0.00987547589465976, 0.18473206134513, 0.405786362942308, 0.331661936361343, 0.00716986577026546, 0.866599955130368, 0.159300076542422, 0.437274595955387, 0.966125315288082, 0.519187876954675, 0.508360112085938, 0.761747567448765, 0.545280912192538, 0.91436234372668, 0.144920523511246, 0.277466043829918, 0.472333801677451, 0.120181784033775, 0.70970232388936, 0.968175527639687, 0.984372295904905, 0.944179642247036, 0.848393188789487, 0.494843865511939, 0.285771421855316, 0.361373612424359, 0.465846951352432, 0.27826995938085, 0.0891039092093706, 0.837274834513664, 0.213070223107934, 0.23897297703661, 0.461390371434391, 0.862498725298792, 0.562024671118706, 0.391832683468238, 0.302512209163979, 0.897878392133862, 0.173774293391034, 0.869689374230802, 0.914288758998737, 0.810755289159715, 0.351586169097573, 0.835025649750605, 0.0800349276978523, 0.37748827249743, 0.401959431124851, 0.783321087947115, 0.897256175754592, 0.232108365977183, 0.625635523814708, 0.459221909986809, 0.581735792336985, 0.619759078137577, 0.509890404297039, 0.452689453959465, 0.244456903310493, 0.402518835151568, 0.430818595690653, 0.539258769713342, 0.66612427867949, 0.302904561860487, 0.414291412802413, 0.961631465470418, 0.0182414785958827, 0.645268196240067, 0.272050599101931, 0.808671454666182, 0.00236757262609899, 0.793411359423772, 0.311784951481968, 0.294548445148394, 0.106361700687557, 0.297042758436874, 0.11741789849475, 0.632668615784496, 0.0379525790922344, 0.796040977118537, 0.647910719970241, 0.823675409192219, 0.530880636768416, 0.824462347198278, 0.506305319489911, 0.239447255153209, 0.0668257467914373, 0.759303470142186, 0.15468867868185, 0.0403608880005777, 0.979374313028529, 0.531115389661863, 0.272600170224905, 0.885529773775488, 0.507740444038063, 0.895402638241649, 0.670111134415492, 0.445800701389089, 0.214784431271255, 0.701798497932032, 0.525804007891566, 0.432907519163564, 0.0529756743926555, 0.646493815118447, 0.821557789808139, 0.727492948295549, 0.630351984873414, 0.403648980893195, 0.802128209266812, 0.360997896175832, 0.0236063548363745, 0.0776362235192209, 0.535060698864982, 0.0768045936711133, 0.185974872205406, 0.698212585179135, 0.156234973808751, 0.757747227791697, 0.171949858777225, 0.409852943383157, 0.0890322299674153, 0.615326184080914, 0.842994191916659, 0.950795059558004, 0.904575115069747, 0.349484382197261, 0.871198021573946, 0.467496300581843, 0.509910311782733, 0.00916957668960094, 0.271083067636937, 0.106755788438022, 0.0346294054761529, 0.257024967111647, 0.055518604349345, 0.803664625855163, 0.68535111262463, 0.217024663928896, 0.925926285330206, 0.30938734090887, 0.808848770568147, 0.325373937375844, 0.920225642621517, 0.507084766635671, 0.675524623598903, 0.992159812012687, 0.35450574173592, 0.935889856191352, 0.602668085601181, 0.54754299740307, 0.814722770359367, 0.748121949844062, 0.554703073576093, 0.316027441760525, 0.725004548905417, 0.307722728000954, 0.492239695973694, 0.983701738528907, 0.0678749573417008, 0.119883633684367, 0.345429417677224, 0.90402725408785, 0.884898557094857, 0.276945462450385, 0.543640675256029, 0.943393583642319, 0.494299709564075, 0.839549085590988, 0.833882945589721, 0.677573835244402, 0.409808993339539, 0.957969944691285, 0.672087891027331, 0.881167553598061, 0.282309855800122, 0.982469100737944, 0.458534933859482, 0.640989337814972, 0.0828128813300282, 0.11393015505746, 0.521995987044647, 0.00736321392469108, 0.0370237366296351, 0.618480717763305, 0.409821235807613, 0.0734905886929482, 0.819596677552909, 0.99974687048234, 0.845833597704768, 0.474555029533803, 0.464849325362593, 0.160016988404095, 0.433911127969623, 0.247177362907678, 0.313642336754128, 0.36173075158149, 0.883852460421622, 0.950986709911376, 0.468700120225549, 0.755599937401712, 0.828293530270457, 0.0990230233874172, 0.112067550886422, 0.631531072081998, 0.683622417505831, 0.401729244040325, 0.500041889026761, 0.479864412918687, 0.121587849920616, 0.827355361776426, 0.129809440812096, 0.000249484088271856, 0.497760427650064, 0.544278095941991, 0.743265227647498, 0.00165420398116112, 0.14520109561272, 0.289012330351397, 0.648584326496348, 0.700882050208747, 0.529250143095851, 0.0966166651342064, 0.188649408286437, 0.144418845884502, 0.101584123447537, 0.673547914018854, 0.467406001640484, 0.516493552131578, 0.562869421206415, 0.936729700770229, 0.825088274199516, 0.991778492694721, 0.5799987136852, 0.650744112906978, 0.729177080327645, 0.335935364477336, 0.0016349065117538, 0.627097368240356, 0.847125875530764, 0.605676477309316, 0.374154339544475, 0.10698433383368, 0.6298418934457, 0.343317817430943, 0.574386265128851, 0.570331123657525, 0.0305428665596992, 0.033089502248913, 0.714553254190832, 0.784701535943896, 0.895686107221991, 0.669425976695493, 0.86930990125984, 0.421519441762939), actual = c(0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"