Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.3454178        0.3507452 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4085962        0.4085962        0.5199843 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.813225448830053, 0.650718489428982, 0.599962769076228, 0.206147171091288, 0.575084319803864, 0.847753146197647, 0.911371055291966, 0.389843423152342, 0.266108844196424, 0.176489586709067, 0.771325343288481, 0.352930813329294, 0.847954114666209, 0.982756863115355, 0.858748102327809, 0.819345608353615, 0.340060706948861, 0.794272139668465, 0.502346261870116, 0.330998349003494, 0.180869634496048, 0.368082076776773, 0.320973284076899, 0.0106753502041101, 0.216185493394732, 0.173753251088783, 0.215232629096135, 0.423074953025207, 0.0299524285364896, 0.520075988257304, 0.199913407675922, 0.889261176111177, 0.0146381533704698, 0.216582132969052, 0.0281004935968667, 0.256237556226552, 0.0509881565812975, 0.128734247293323, 0.462594935437664, 0.800918813096359, 0.621389496838674, 0.199967303778976, 0.172136744251475, 0.707943903282285, 0.33088794304058, 0.886748051270843, 0.246193679748103, 0.0431491476483643, 0.673222355078906, 0.743480402743444, 0.790111594134942, 0.617135452106595, 0.873247161973268, 0.544620632426813, 0.132837390061468, 0.785817354219034, 0.641111637465656, 0.704273394541815, 0.688198385061696, 0.565661967964843, 0.677428535418585, 0.266455069184303, 0.378101041540504, 0.428957316325977, 0.190812294371426, 0.153658471303061, 0.455581129994243, 0.500639417907223, 0.843218073248863, 0.539305767277256, 0.697564910398796, 0.313390588155016, 0.350086522754282, 0.360522189177573, 0.263525102986023, 0.374625651398674, 0.0149302447680384, 0.697751634288579, 0.400648797862232, 0.769509783713147, 0.605503124650568, 0.317716352641582, 0.315543489763513, 0.903907873434946, 0.775878824526444, 0.247952685691416, 0.694460518658161, 0.0226588288787752, 0.138956453185529, 0.615851725218818, 0.901671650120988, 0.715379402972758, 0.0849635461345315, 0.32048553833738, 0.790954180061817, 0.949655099306256, 0.225961251184344, 0.608442544937134, 0.683062616270036, 0.126216380624101, 0.44114310760051, 0.899305785307661, 0.745161343133077, 0.397276101168245, 0.901519156759605, 0.599459205986932, 0.810682496521622, 0.381694980431348, 0.108296690974385, 0.803673649905249, 0.490836758865044, 0.971023243619129, 0.60791686293669, 0.75393517059274, 0.784072291571647, 0.708633978385478, 0.79178215120919, 0.223960454575717, 0.995229568798095, 0.77421482373029, 0.205871337791905, 0.259486535331234, 0.0861871915403754, 0.245294495485723, 0.870552134700119, 0.534502918832004, 0.1696672225371, 0.240788030903786, 0.154281800379977, 0.731301145860925, 0.9498738842085, 0.980052822269499, 0.943037773482502, 0.190120344050229, 0.774287656182423, 0.136031712638214, 0.802727166563272, 0.35450598387979, 0.381156813120469, 0.258252140600234, 0.561842687427998, 0.257678705733269, 0.25021786428988, 0.548783841077238, 0.636876635020599, 0.425204393221065, 0.000977500109001994, 0.45615626918152, 0.137548791477457, 0.666089066304266, 0.897466687951237, 0.12509767152369, 0.548032694030553, 0.950503409840167, 0.79750832519494, 0.684477263595909, 0.363210263894871, 0.428682732628658, 0.172976035857573, 0.219186699483544, 0.742075142450631, 0.757946442114189, 0.66847876412794, 0.172507451381534, 0.547673641005531, 0.483421264449134, 0.392802946502343, 0.229313912335783, 0.769565319642425, 0.723600610159338, 0.980470065027475, 0.783853044267744, 0.805002439301461, 0.243383131688461, 0.230466109933332, 0.955909985117614, 0.547040966106579, 0.324887639144436, 0.150558612542227, 0.0775193937588483, 0.371861258521676, 0.917614348232746, 0.520681383553892, 0.715010233223438, 0.208994987886399, 0.699966527987272, 0.3762520221062, 0.0217627387028188, 0.246590655995533, 0.224177701864392, 0.583890154724941, 0.973646736005321, 0.163013465935364, 0.520844166632742, 0.645545470295474, 0.965317005058751, 0.900225358316675, 0.622357947519049, 0.124455857090652, 0.703617737162858, 0.51966478349641, 0.612683225888759, 0.218988278415054, 0.530340739758685, 0.915852191857994, 0.873985170153901, 0.0963837809395045, 0.89379810821265, 0.594471068121493, 0.611476429505274, 0.749839675845578, 0.0636498921085149, 0.0545902738813311, 0.336284971795976, 0.449184057302773, 0.405661170138046, 0.220917963888496, 0.437201255233958, 0.180892296135426, 0.111739017767832, 0.858256237115711, 0.918566505657509, 0.0110588672105223, 0.674012137111276, 0.311692563584074, 0.728078054729849, 0.841551067074761, 0.541988536017016, 0.244649557163939, 0.484104366973042, 0.77537199202925, 0.0541831431910396, 0.201065406668931, 0.106322581181303, 0.54046000004746, 0.816794055281207, 0.633637746563181, 0.142364287748933, 0.945755661698058, 0.549656926654279, 0.717903345357627, 0.844512419775128, 0.427341655595228, 0.29205630114302, 0.168793309945613, 0.303615463199094, 0.855266294442117, 0.147739319130778, 0.583333138842136, 0.859811802860349, 0.789235087111592, 0.615256424294785, 0.084161693463102, 0.37492955615744, 0.24211478070356, 0.166399583918974, 0.95469307503663, 0.68505496205762, 0.112617388833314, 0.582112538861111, 0.0424985990393907, 0.781990151619539, 0.973410275299102, 0.595754412468523, 0.848481433466077, 0.485325000016019, 0.260584187461063, 0.37019388563931, 0.304594983579591, 0.680017196806148, 0.795976468594745, 0.770346807548776, 0.140506222145632, 0.208971321117133, 0.914012488676235, 0.8061724815052, 0.345334295649081, 0.558141758199781, 0.439772001933306, 0.336260873125866, 0.317434340482578, 0.603317134780809, 0.39537592465058, 0.676461128750816, 0.346128046745434, 0.804598847171292, 0.203652260592207, 0.530971751315519, 0.996570025105029, 0.257068604230881, 0.117257081437856, 0.300935194594786, 0.744420566596091, 0.181064443895593, 0.441070269793272, 0.607405521674082, 0.45982871693559, 0.368896827334538, 0.112733000889421, 0.274233829462901, 0.606064349412918, 0.0388788532000035, 0.0681246223393828, 0.546487409621477, 0.530715130502358, 0.421097392681986, 0.430582782952115, 0.649835477117449, 0.581344791222364, 0.55715338839218, 0.804441704647616, 0.29727016040124, 0.629768377402797, 0.77500463440083, 0.79826488904655, 0.976869034348056, 0.409335685195401, 0.278054223628715, 0.613058238988742, 0.294115000637248, 0.113739472115412, 0.886951220920309, 0.869639215292409, 0.0435287107247859, 0.627014762489125, 0.671803731238469, 0.297011704416946, 0.0716838794760406, 0.272277009906247, 0.157741939183325, 0.3405183237046, 0.279644445283338, 0.628706797491759, 0.582802855176851, 0.156393276294693, 0.891855935566127, 0.707657305523753, 0.391540622105822, 0.0640026708133519, 0.319614720530808, 0.907773264683783, 0.859898965572938, 0.948066964512691, 0.808383404044434, 0.258346195798367, 0.474084215937182, 0.981124552898109, 0.943097905255854, 0.507802553009242, 0.783117251237854, 0.266132886288688, 0.559742208803073, 0.293541675200686, 0.54816728993319, 0.869782508583739, 0.00158073566854, 0.541278643999249, 0.98297255858779, 0.141822871053591, 0.383876534178853, 0.588602139381692, 0.741048178402707, 0.081925505772233, 0.574387619970366, 0.520516893360764, 0.975724875926971, 0.60964567004703, 0.576427154941484, 0.513317951932549, 0.102522616973147, 0.780889322748408, 0.668244512751698, 0.369926364626735, 0.635535379871726, 0.131406325148419, 0.335054745199159, 0.731285484740511, 0.743231675587595, 0.755147431744263, 0.340990775031969, 0.13727972493507, 0.0222137251403183, 0.484824259765446, 0.130527684697881, 0.759995869128034, 0.772054037777707, 0.782037858152762, 0.427585944533348, 0.7072531983722, 0.55032068816945, 0.758324551861733, 0.894689205102623, 0.747943509137258, 0.605940074194223, 0.137228396022692, 0.358783395029604, 0.711308092810214, 0.500133669003844, 0.62488075136207, 0.0297931081149727, 0.428372408496216, 0.992827454581857, 0.766155383083969, 0.46771496348083, 0.740962481824681, 0.117433301173151, 0.722183130914345, 0.426424195524305, 0.625978991622105, 0.465627373661846, 0.01971366093494, 0.197352034272626, 0.884788275230676, 0.491770623950288, 0.258747531333938, 0.0230383705347776, 0.596782042644918, 0.237727470928803, 0.899814732605591, 0.118399365339428, 0.0781312314793468, 0.0916658290661871, 0.398124560480937, 0.389808270148933, 0.725491384509951, 0.460302893305197, 0.910043791867793, 0.193834885023534, 0.596501685678959, 0.597765144659206, 0.957201192621142, 0.0661890651099384, 0.312301906989887, 0.115640290779993, 0.162211816292256, 0.236390353180468, 0.815885584102944, 0.259010113310069, 0.99529464263469, 0.584477643715218, 0.856542598921806, 0.277405423577875, 0.465132946381345, 0.659226263407618, 0.0158575179520994, 0.15577856102027, 0.0666913231834769, 0.817827174207196, 0.439726800657809, 0.0899569909088314, 0.207028031582013, 0.346566333668306, 0.573283599456772, 0.20929406536743, 0.674701075535268, 0.250280103879049, 0.601661065127701, 0.199499992653728, 0.0684143514372408, 0.821783397346735, 0.809155219933018, 0.634485584916547, 0.343292218632996, 0.0233025914058089, 0.545792904682457, 0.0804903593380004, 0.104749298887327, 0.548004408832639, 0.418627694947645, 0.0554946183692664, 0.701004394097254, 0.328526194905862, 0.0788008205126971, 0.146878640633076, 0.101916352985427, 0.673428294947371, 0.881514214444906, 0.156463816761971, 0.664417503867298, 0.680780436610803, 0.739321702392772, 0.687922704964876, 0.435058315517381, 0.407134997891262, 0.776670941151679, 0.563203451456502, 0.88728466280736, 0.0644395060371608, 0.390467578312382, 0.37664495408535, 0.144684629747644, 0.902671824442223, 0.16026800358668, 0.213254689704627, 0.179334006039426, 0.538449025014415, 0.430907518835738, 0.0635599137749523, 0.718739236937836, 0.116315019782633, 0.109156711259857, 0.720861037960276, 0.00301258801482618, 0.287757520796731, 0.436770936707035, 0.600817405851558, 0.624750061891973, 0.199852183461189, 0.805256560212001, 0.649475252255797, 0.531645027920604, 0.534703274490312, 0.538677287753671, 0.977105093887076, 0.0634469017386436, 0.334377748193219, 0.735422033350915, 0.269895625300705, 0.545433474238962, 0.382563630351797, 0.816470520105213, 0.0168576538562775, 0.274685082025826, 0.345462589757517, 0.0877194660715759, 0.326183147029951, 0.487928337417543, 0.433831098955125, 0.196969720767811, 0.0863768404815346, 0.679746233625337, 0.669581927126274, 0.266434105811641, 0.220348244998604, 0.997739404207096, 0.58552504889667, 0.537672399776056, 0.467464081244543, 0.461756668053567, 0.00367162609472871, 0.169689838308841, 0.721459725406021, 0.598704156465828, 0.964519231347367, 0.0835032516624779, 0.575815640157089, 0.296948852017522, 0.165439749136567, 0.653029255801812, 0.559880731161684, 0.659353557508439, 0.269895405741408, 0.0256550181657076, 0.151291613467038, 0.487601523986086, 0.758352927863598, 0.712163584539667, 0.42376139620319, 0.462799861328676, 0.163289390038699, 0.364151702262461, 0.407932159258053, 0.831573688425124, 0.989938982762396, 0.43726704432629, 0.743002906674519, 0.773601722437888, 0.730428061680868, 0.868647335097194, 0.25694977492094, 0.377682771533728, 0.422378807561472, 0.840509931091219, 0.310196093982086, 0.285311808111146, 0.829025330720469, 0.56523658381775, 0.479389375308529, 0.261802851688117, 0.172622363083065, 0.582116795936599, 0.888533237855881, 0.799296989571303, 0.29497711895965, 0.886087599676102, 0.360042593674734, 0.108706155791879, 0.344642077805474, 0.578986229375005, 0.292014719452709, 0.615888903150335, 0.304105039453134, 0.815082705346867, 0.64977277466096, 0.288307112874463, 0.410503621445969, 0.849022542359307, 0.39003851567395, 0.580186000093818, 0.997564211953431, 0.538775755790994, 0.802112588426098, 0.162796539487317, 0.21155890985392, 0.946922404225916, 0.770389292389154, 0.325308570405468, 0.928215674357489, 0.941792411496863, 0.0446081829722971, 0.556785058695823, 0.555001392960548, 0.547914414433762, 0.61712699290365, 0.657775011379272, 0.84547436540015, 0.945896480465308, 0.0167779421899468, 0.422689435305074, 0.184445700375363, 0.80579468072392, 0.338380825472996, 0.901414847699925, 0.0359218425583094, 0.875277556246147, 0.742239977698773, 0.881795953959227, 0.474686363013461, 0.792087966809049, 0.730339916655794, 0.556881348602474, 0.233149593230337, 0.483288746094331, 0.159207477234304, 0.524610383901745, 0.619786231545731, 0.264564678771421, 0.0261446961667389, 0.198581453412771, 0.00597171974368393, 0.158101146575063, 0.81985683995299, 0.747533406596631, 0.33268303819932, 0.478267552796751, 0.284681425662711, 0.62656211713329, 0.925947083160281, 0.538724059937522, 0.825556312687695, 0.667510265717283, 0.588652637554333, 0.259464549366385, 0.123817759566009, 0.0348125693853945, 0.74938499648124, 0.816390603315085, 0.71506466018036, 0.193169386824593, 0.2013557148166, 0.123367271618918, 0.315151496324688, 0.771144622936845, 0.953304235823452, 0.830425299005583, 0.420366676524282, 0.497732912423089, 0.894032356329262, 0.561287237098441, 0.267485504504293, 0.718551072757691, 0.894581237342209, 0.20543483668007, 0.0374007776845247, 0.532986238133162, 0.423815440153703, 0.755858606193215, 0.687617271440104, 0.970698482822627, 0.739000516245142, 0.614418487530202, 0.202649803599343, 0.971599007956684, 0.489759593270719, 0.348143299110234, 0.641084598377347, 0.151097990339622, 0.125155880348757, 0.724587631644681, 0.0522041779477149, 0.378701883135363, 0.44156724284403, 0.756949981208891, 0.484761754749343, 0.728374249301851, 0.0273321026470512, 0.716582264751196, 0.556651424616575, 0.383817149559036, 0.481507031945512, 0.702769136289135, 0.315563094103709, 0.310723870526999, 0.988262387923896, 0.190363881876692, 0.595694934250787, 0.554976032814011, 0.322012388380244, 0.663473080145195, 0.0834984318353236, 0.0494043144863099, 0.0615337875206023, 0.189177389256656, 0.878230873961002, 0.376584828598425, 0.61995688918978, 0.587369965389371, 0.282826126553118, 0.899188657058403, 0.487216793466359, 0.390542260603979, 0.983038058271632, 0.681064951699227, 0.301821129629388, 0.338222751626745, 0.745704896049574, 0.780540346633643, 0.0865731723606586, 0.0277798546012491, 0.361616661539301, 0.260046287439764, 0.187889315420762, 0.162800652440637, 0.85120303183794, 0.713763275649399, 0.626519654178992, 0.0252442450728267, 0.739998410455883, 0.702165234833956, 0.513825529254973, 0.523970745969564, 0.76651475043036, 0.513968650717288, 0.882040174910799, 0.127253681886941, 0.49617447052151, 0.19526629964821, 0.303674675524235, 0.793402284150943, 0.238815506687388, 0.397878230549395, 0.956643158337101, 0.0512284750584513, 0.344468067167327, 0.157358836615458, 0.705856973305345, 0.764223640784621, 0.183942677685991, 0.20720522897318, 0.947078763740137, 0.600683814613149, 0.144374442286789, 0.034390302374959, 0.204497823491693, 0.332422100240365, 0.618728836067021, 0.133895517326891, 0.374742400366813, 0.537103531649336, 0.496198322856799, 0.732262515695766, 0.511596633354202, 0.197847747942433, 0.0966595492791384, 0.794747381936759, 0.180062837200239, 0.846924826269969, 0.712666964856908, 0.859962468035519, 0.819293533219025, 0.416357595007867, 0.361924971221015, 0.047486568801105, 0.688990162918344, 0.899149952456355, 0.9755797625985, 0.609294188441709, 0.510157502954826, 0.611428293632343, 0.546548846643418, 0.729767237091437, 0.842890070285648, 0.712434598244727, 0.174756644060835, 0.826823577051982, 0.803497474407777, 0.00893392530269921, 0.609897633548826, 0.948242617072538, 0.582358394749463, 0.414565929211676, 0.352527167182416, 0.558633802691475, 0.985090411733836, 0.684035894460976, 0.0729326866567135, 0.257483385968953, 0.0829072166234255, 0.649335928959772, 0.745622685877606, 0.92019419092685, 0.824960472527891, 0.358759975526482, 0.840833365218714, 0.14485974307172, 0.60215606726706, 0.383537327172235, 0.624966513598338, 0.258843559771776, 0.932356920093298, 0.822312219999731, 0.900086462497711, 0.409932477399707, 0.890162509633228, 0.149068454746157, 0.996117333648726, 0.110269708326086, 0.87286817189306, 0.28126646229066, 0.839371782261878, 0.349858449306339, 0.330190303269774, 0.415264102863148, 0.963342529954389, 0.372090361313894, 0.89152566366829, 0.790463944431394, 0.521726106759161, 0.33335595834069, 0.845950554125011, 0.10462581156753, 0.139078145148233, 0.15320757497102, 0.455703135114163, 0.219987041549757, 0.360631836112589, 0.542772183427587, 0.318664540536702, 0.790278293425217, 0.468531025806442, 0.0842570741660893, 0.321287347702309, 0.818938059499487, 0.674831426469609, 0.592012672219425, 0.654138329206035, 0.167318020714447, 0.488051944645122, 0.277956035220996, 0.622400103369728, 0.634086753008887, 0.236905874684453, 0.473832795862108, 0.409578983671963, 0.838148988084868, 0.756762483622879, 0.108157102717087, 0.147108280099928, 0.0908308359794319, 0.7622479069978, 0.0111873752903193, 0.0918645306956023, 0.465896877693012, 0.527931058313698, 0.878837064374238, 0.786889642709866, 0.0431871670298278, 0.775064826244488, 0.115172340068966, 0.328366689383984, 0.869359024101868, 0.118589900666848, 0.257503596134484, 0.137520599877462, 0.919288412900642, 0.35446464526467, 0.580281760543585, 0.515643876744434, 0.432531706057489, 0.88662481540814, 0.130699081113562, 0.407105990452692, 0.340211663395166, 0.258443140890449, 0.60136837977916, 0.981183480704203, 0.712464213604107, 0.269861843669787, 0.025207496015355, 0.59479584172368, 0.108103818260133, 0.413849125150591, 0.928993011126295, 0.375477348687127, 0.926978891948238, 0.381755941780284, 0.25337232183665, 0.269107793923467, 0.258518640184775, 0.997786173131317, 0.305630146991462, 0.539294152054936, 0.648923244560137, 0.287014578003436, 0.203848880715668, 0.0963097072672099, 0.652506930287927, 0.0366501442622393, 0.840005650417879, 0.440349900396541, 0.055610197596252, 0.253574929432943, 0.887812461238354, 0.232306187506765, 0.907158276997507, 0.761701984331012, 0.325150784337893, 0.512751458212733, 0.57498759101145, 0.0394825153052807, 0.759010016452521, 0.369277460733429, 0.284531533019617, 0.982598948059604, 0.235207224264741, 0.469635650049895, 0.322358704172075, 0.397930561332032, 0.801121592056006, 0.229883134830743, 0.469604545738548, 0.524890358559787, 0.0715352927800268, 0.491091877920553, 0.596711463760585, 0.63021966163069, 0.238485343987122, 0.499825075967237, 0.218320957617834, 0.773669084534049, 0.657694739056751, 0.474481550045311, 0.0156664946116507, 0.465382418362424, 0.541932608699426, 0.154870923608541, 0.146077168174088, 0.293858899967745, 0.0201098737306893, 0.511233073892072, 0.971684524323791, 0.370877034962177, 0.861528710229322, 0.980234926566482, 0.990613938542083, 0.323323806747794, 0.859900797950104, 0.665390075417235, 0.0379026171285659, 0.644394481554627, 0.943717230111361, 0.220484078628942, 0.520048344507813, 0.515654855407774, 0.261108675971627, 0.218165196711197, 0.831949346000329, 0.341420128010213, 0.816172647057101, 0.212968711042777, 0.93315236736089, 0.422398750437424, 0.232995800906792, 0.656013670377433, 0.0822251655627042, 0.191571714589372, 0.0483130072243512, 0.629096518270671, 0.0810955932829529, 0.0300774346105754, 0.980142907937989, 0.724551724269986, 0.871036663185805, 0.340430126292631, 0.447064234409481, 0.150079693179578, 0.882704597897828, 0.00614480837248266, 0.631435796385631, 0.83410738571547), actual = c(0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.422080618562177, 0.627176546258852, 0.43961840076372, 0.32585196569562, 0.811067651258782, 0.710219486849383, 0.0669541019015014, 0.0215076962485909, 0.0724081178195775, 0.443828379269689, 0.852562614483759, 0.261026879074052, 0.737008190480992, 0.482641296694055, 0.446582492208108, 0.581692603649572, 0.235467428341508, 0.331469144206494, 0.0364134162664413, 0.337092502741143, 0.496925046434626, 0.732824056409299, 0.372544188518077, 0.5726439722348, 0.242594552226365, 0.425838242052123, 0.59701678249985, 0.203410378890112, 0.467531491536647, 0.393843384459615, 0.946559532079846, 0.366502283606678, 0.598739647539333, 0.474567330442369, 0.0024170211981982, 0.310582962352782, 0.0733876042068005, 0.491389342583716, 0.520132519304752, 0.436713972361758, 0.432760819792747, 0.243531279964373, 0.913484980817884, 0.350999489892274, 0.86939204717055, 0.374919947935268, 0.309291412122548, 0.799059685552493, 0.572624409338459, 0.933222834719345, 0.504964996594936, 0.217845876934007, 0.682802203577012, 0.564437227323651, 0.366237965645269, 0.389966159593314, 0.942481503356248, 0.696726836962625, 0.847292275400832, 0.993213697802275, 0.3910644329153, 0.274598548887298, 0.715072489576414, 0.426138385431841, 0.63340019620955, 0.903414150932804, 0.00699363159947097, 0.430723218014464, 0.9570020979736, 0.969930109800771, 0.45442643831484, 0.531875847605988, 0.520634394139051, 0.805066821398214, 0.635770083405077, 0.48320314870216, 0.887872356688604, 0.749888502061367, 0.400007556425408, 0.628442264860496, 0.68492262577638, 0.382350735599175, 0.132642104057595, 0.092013540212065, 0.00910351588390768, 0.699522094568238, 0.298407291527838, 0.0622245646081865, 0.985810616984963, 0.572890547802672, 0.58325294079259, 0.504340365761891, 0.515541126020253, 0.128332366468385, 0.434733518864959, 0.0455520050600171, 0.659226044313982, 0.369104558834806, 0.957571977283806, 0.352900993311778, 0.25526335905306, 0.639315761858597, 0.617669468047097, 0.0391222825273871, 0.198247305583209, 0.998543621739373, 0.337912784190848, 0.597564137075096, 0.348688417812809, 0.72778023709543, 0.419471209403127, 0.215943252900615, 0.803143405355513, 0.0390911013819277, 0.229339133249596, 0.778725378448144, 0.914787276415154, 0.136391981272027, 0.490522286621854, 0.722516076173633, 0.494082290446386, 0.0569394514895976, 0.991262380732223, 0.00703976722434163, 0.110820937668905, 0.879469779087231, 0.993132523959503, 0.422983509721234, 0.958462109323591, 0.574836459709331, 0.57179535436444, 0.237907974747941, 0.306468568276614, 0.327881173696369, 0.275731450645253, 0.634246797999367, 0.837196259759367, 0.90585611294955, 0.324985920451581, 0.335426114499569, 0.887570006540045, 0.159261858323589, 0.0234702376183122, 0.599374281708151, 0.962459777016193, 0.543959436705336, 0.760642592795193, 0.435564241604879, 0.862437279429287, 0.907283136853948, 0.481376859592274, 0.40350152249448, 0.0783680053427815, 0.281577841378748, 0.447439290815964, 0.0471402157563716, 0.786428189137951, 0.0165433785878122, 0.714047077810392, 0.283465696498752, 0.551039146492258, 0.87983577651903, 0.450706965988502, 0.663540943991393, 0.0909331957809627, 0.132782757747918, 0.986902649747208, 0.571720710489899, 0.902743073413149, 0.676627196837217, 0.629606892121956, 0.156445400556549, 0.545542798005044, 0.675863125128672, 0.135121054947376, 0.25785121275112, 0.45895920926705, 0.825165703892708, 0.418634257977828, 0.0818368603941053, 0.651622967328876, 0.984937537228689, 0.882855322677642, 0.549630905734375, 0.194875625427812, 0.876799736637622, 0.0443449190352112, 0.763295154320076, 0.817887259880081, 0.697553908685222, 0.508727835956961, 0.509606754407287, 0.0364209399558604, 0.0767459594644606, 0.797861301572993, 0.115173072787002, 0.140825319103897, 0.582420648541301, 0.15034005488269, 0.976456116419286, 0.492519387044013, 0.545067253755406, 0.918993768282235, 0.446335965767503, 0.76592337526381, 0.977441407740116, 0.958922063698992, 0.872644542017952, 0.188759937416762, 0.912216139957309, 0.896801989525557, 0.798454033210874, 0.151820787927136, 0.629447501851246, 0.13508415245451, 0.235566400690004, 0.664390082703903, 0.922477060230449, 0.726673104334623, 0.452111922204494, 0.271122455131263, 0.821936492808163, 0.436943442560732, 0.502848423318937, 0.468190879095346, 0.491112918127328, 0.738360335119069, 0.799779965309426, 0.648340070620179, 0.409215381834656, 0.150593588594347, 0.72821557149291, 0.780142020666972, 0.476080761756748, 0.0506523526273668, 0.646704723592848, 0.487445133505389, 0.742261262377724, 0.228016604203731, 0.885841796640307, 0.842757290927693, 0.114541074726731, 0.613130821380764, 0.153529373928905, 0.743145293323323, 0.352402329444885, 0.325179712148383, 0.177060626912862, 0.230580791831017, 0.535501933656633, 0.305288805393502, 0.247055262094364, 0.75737469480373, 0.97542197490111, 0.866540306247771, 0.6087045755703, 0.419023004826158, 0.300988312577829, 0.65189933241345, 0.614637456601486, 0.441587223438546, 0.207741309423, 0.397928184829652, 0.912973801605403, 0.40146243898198, 0.972421003505588, 0.419995589647442, 0.330494611058384, 0.293157369364053, 0.669525141362101, 0.631919692037627, 0.527442200575024, 0.083932600915432, 0.15517926774919, 0.539413079852238, 0.481036149198189, 0.489435723051429, 0.139650724129751, 0.849913732847199, 0.801699644187465, 0.746277789352462, 0.929218001198024, 0.102408022386953, 0.0930011586751789, 0.736792832380161, 0.666528721340001, 0.880875051952899, 0.89358735922724, 0.341347691603005, 0.206665713805705, 0.498247431125492, 0.876039994182065, 0.289392448728904, 0.45063776592724, 0.813818626105785, 0.487814174732193, 0.621442706091329, 0.601036124862731, 0.408667409559712, 0.383242318406701, 0.518622153438628, 0.754222253104672, 0.874348519137129, 0.335461471462622, 0.850525843678042, 0.219995370134711, 0.35617799195461, 0.0426773554645479, 0.830225799465552, 0.089694473426789, 0.14019187213853, 0.434682722669095, 0.706385721452534, 0.777301491936669, 0.116185929859057, 0.830384529661387, 0.0484440794680268, 0.403461458161473, 0.110289306147024, 0.843892879784107, 0.157509078970179, 0.9009604442399, 0.372022681171075, 0.0818301916588098, 0.896910701412708, 0.0278345162514597, 0.839553711004555, 0.859341998351738, 0.688285637181252, 0.951865588547662, 0.195662735495716, 0.528264068998396, 0.820680975914001, 0.646099332021549, 0.521483073011041, 0.00122184259817004, 0.766306483652443, 0.291406060801819, 0.864822952542454, 0.179326939629391, 0.927714040968567, 0.809501499403268, 0.899752430384979, 0.411725549260154, 0.518316112924367, 0.950973422266543, 0.218854180071503, 0.967764821602032, 0.730080610606819, 0.285420927451923, 0.500208440236747, 0.279496283037588, 0.834644386079162, 0.719974570907652, 0.415233181789517, 0.516208741813898, 0.242702735820785, 0.548372536664829, 0.644177564419806, 0.259482456138358, 0.25259946170263, 0.578030880074948, 0.593547827797011, 0.688510409323499, 0.772855714894831, 0.595853845821694, 0.341068107169122, 0.397743169451132, 0.413414775859565, 0.944528131512925, 0.909547326155007, 0.477897743927315, 0.00563239166513085, 0.507693289313465, 0.485504187876359, 0.244284824235365, 0.671926779905334, 0.127654270734638, 0.266144277062267, 0.984424445778131, 0.386362922145054, 0.845188252395019, 0.33910023374483, 0.0758169619366527, 0.0727768761571497, 0.992103384342045, 0.554756443249062, 0.940211383160204, 0.109470120165497, 0.753025522222742, 0.671938323881477, 0.362356354948133, 0.751243773847818, 0.863820203579962, 0.701669763075188, 0.693527844268829, 0.294828708283603, 0.263486797222868, 0.920259122736752, 0.522718141321093, 0.97977288486436, 0.0938528680708259, 0.448811116162688, 0.467205176595598, 0.661266030743718, 0.118633599719033, 0.195486497599632, 0.196649687597528, 0.455603126203641, 0.787536468124017, 0.409779000096023, 0.555229055462405, 0.916183194378391, 0.682960860664025, 0.0356527199037373, 0.415446314262226, 0.234907190082595, 0.822807357646525, 0.589891017880291, 0.129629229428247, 0.281023852992803, 0.563860540045425, 0.173803024692461, 0.332497707102448, 0.705569081706926, 0.436061787651852, 0.306188876973465, 0.892040300182998, 0.0835351501591504, 0.368417968507856, 0.110086356755346, 0.288487787358463, 0.320101173594594, 0.949725135462359, 0.796299118548632, 0.373316464014351, 0.975938645657152, 0.373447984457016, 0.792346836533397, 0.147636615904048, 0.298037776956335, 0.231860019266605, 0.821132569108158, 0.314563108608127, 0.337543838657439, 0.389062989503145, 0.674905901076272, 0.575103923678398, 0.00814288831315935, 0.198329370701686, 0.1243158983998, 0.505616217153147, 0.555214853724465, 0.742604953702539, 0.943466218188405, 0.0902106880676001, 0.133184392703697, 0.975374144036323, 0.798602519324049, 0.162500179605559, 0.247426748275757, 0.183206626446918, 0.371512930840254, 0.424865341279656, 0.640832515666261, 0.211529729655012, 0.180568234529346, 0.398921792861074, 0.649918939219788, 0.340318856062368, 0.674156315391883, 0.933271294459701, 0.274222071515396, 0.0588149833492935, 0.276311087189242, 0.680907275993377, 0.299857407109812, 0.497390836244449, 0.0963296389672905, 0.342045887839049, 0.0360047211870551, 0.911130144726485, 0.855333481682464, 0.0488519330974668, 0.00889911572448909, 0.164562186226249, 0.474788035731763, 0.0553392132278532, 0.609512317692861, 0.383095502387732, 0.0465541086159647, 0.125225317431614, 0.849047165596858, 0.378289165673777, 0.161525833886117, 0.551781146787107, 0.151235591620207, 0.178798079490662, 0.965300603304058, 0.941832369193435, 0.791986496653408, 0.530382711440325, 0.565772277535871, 0.96318444958888, 0.266727081267163, 0.90095570217818, 0.117576229386032, 0.920344719896093, 0.42719124769792, 0.242463417816907, 0.845933087402955, 0.981785418698564, 0.158014012267813, 0.415276422863826, 0.199645265005529, 0.356670605484396, 0.52688028709963, 0.15391900180839, 0.0339392623864114, 0.748870542738587, 0.673238481162116, 0.543895019916818, 0.307931957766414, 0.419276732252911, 0.265766091179103, 0.814648797968403, 0.505403739633039, 0.155428780941293, 0.887138697551563, 0.532313527539372, 0.577608294785023, 0.242908576270565, 0.672964870696887, 0.333990263752639, 0.103907865937799, 0.510312806349248, 0.0615437058731914, 0.759010252077132, 0.279147992143407, 0.754014238715172, 0.96202535321936, 0.426736922236159, 0.58317905315198, 0.907787061296403, 0.186260510236025, 0.977111329557374, 0.666427572723478, 0.547314906958491, 0.514613314764574, 0.764564286218956, 0.556556793162599, 0.246371668064967, 0.0319178958889097, 0.578965352615342, 0.290778188500553, 0.0187582133803517, 0.112834603525698, 0.344817931065336, 0.853142346022651, 0.942290689330548, 0.0496394999790937, 0.235946476925164, 0.569260357413441, 0.696256547933444, 0.939816917059943, 0.611733876634389, 0.9929562616162, 0.88454618351534, 0.0692254253663123, 0.735888160765171, 0.324893052689731, 0.227190942503512, 0.0268881823867559, 0.208711243001744, 0.748801579931751, 0.600457420339808, 0.779470172477886, 0.408628797158599, 0.236153029603884, 0.436917923158035, 0.773292521946132, 0.061013224767521, 0.699292058125138, 0.992666959995404, 0.80586608313024, 0.0990680805407465, 0.648321075132117, 0.92727141501382, 0.259342261822894, 0.221949033439159, 0.0370577359572053, 0.872754186857492, 0.095813968218863, 0.280316773802042, 0.546034357277676, 0.299848481081426, 0.65782110625878, 0.764306114520878, 0.470308788586408, 0.314711617305875, 0.781819405965507, 0.0371528866235167, 0.5595638663508, 0.586986295180395, 0.698474587872624, 0.312243422260508, 0.454854384297505, 0.310065744211897, 0.232678537489846, 0.903716497588903, 0.0707698995247483, 0.202777813188732, 0.0314006363041699, 0.880101210670546, 0.596200812142342, 0.457056238781661, 0.514682127628475, 0.253089095233008, 0.597176992334425, 0.823936116648838, 0.447727515362203, 0.23858054401353, 0.685551498318091, 0.283411831362173, 0.886863089632243, 0.864567688200623, 0.471426662057638, 0.0298196179792285, 0.863112617051229, 0.0020435752812773, 0.466598988045007, 0.437929413514212, 0.691858436679468, 0.439054828137159, 0.0731953538488597, 0.698763536056504, 0.374001485062763, 0.765911224298179, 0.861865536542609, 0.231187763623893, 0.836747286142781, 0.199902663705871, 0.846914677880704, 0.724208352854475, 0.617631491040811, 0.109496490331367, 0.0861429572105408, 0.721553482813761, 0.24889372754842, 0.442120337393135, 0.629122896818444, 0.373543885769323, 0.274625832214952, 0.245058448053896, 0.809759843861684, 0.22991657233797, 0.871365930419415, 0.395864003105089, 0.679648349992931, 0.0641080527566373, 0.094296922441572, 0.781158114783466, 0.467728830175474, 0.68921660981141, 0.0125575717538595, 0.0120849271770567, 0.604143598815426, 0.763295211829245, 0.207140263402835, 0.612733695423231, 0.0477119379211217, 0.314166206168011, 0.92846266948618, 0.0137327341362834, 0.506980310659856, 0.446608688915148, 0.316889951936901, 0.715252451132983, 0.131963600870222, 0.0174529615323991, 0.19949698029086, 0.507958247559145, 0.0190886766649783, 0.092492196476087, 0.8463169131428, 0.035331369843334, 0.362121800426394, 0.285132175777107, 0.0172129299025983, 0.604948171414435, 0.680563753237948, 0.250398020725697, 0.144356964388862, 0.146338752936572, 0.313224561978132, 0.293715805048123, 0.654256229288876, 0.861047279555351, 0.947817977750674, 0.252142549725249, 0.422650867840275, 0.584348682779819, 0.825995992636308, 0.371663781814277, 0.477508746087551, 0.133304515853524, 0.808463384164497, 0.994074237998575, 0.671809242805466, 0.389498786069453, 0.138154826592654, 0.467028829734772, 0.157741719391197, 0.33329518022947, 0.984342610696331, 0.478550654137507, 0.513883510604501, 0.38011350017041, 0.926190085709095, 0.101624647388235, 0.000423746649175882, 0.659159908303991, 0.0640646922402084, 0.408474198076874, 0.206093647517264, 0.473903307924047, 0.617962518008426, 0.748241960303858, 0.802219078177586, 0.824344424763694, 0.42551650153473, 0.596482230117545, 0.387755439151078, 0.882275824202225, 0.0601805716287345, 0.534983989782631, 0.991993590723723, 0.204134410945699, 0.748474041000009, 0.667729679262266, 0.453604825306684, 0.525879309512675, 0.993354573380202, 0.480036857072264, 0.763247426832095, 0.939842741237953, 0.763322371290997, 0.855259905802086, 0.429121620720252, 0.808444758877158, 0.32121089217253, 0.135587375843897, 0.264545485610142, 0.0515828861389309, 0.870332155842334, 0.017272587865591, 0.27651816373691, 0.167596248909831, 0.947784916730598, 0.0287446093279868, 0.24960745847784, 0.396798040019348, 0.242499768966809, 0.0495443888939917, 0.451554989907891, 0.115512314951047, 0.382477205246687, 0.263742078794166, 0.0680258893407881, 0.322048269677907, 0.480786984553561, 0.405761247035116, 0.000667080748826265, 0.480429958086461, 0.531790842534974, 0.320113799534738, 0.451947647845373, 0.308432574151084, 0.449807208962739, 0.344500828301534, 0.724865901051089, 0.562975717242807, 0.045051799621433, 0.482353263534606, 0.171135724289343, 0.589867880567908, 0.815510970773175, 0.812004894018173, 0.586008572019637, 0.0858549494296312, 0.634876787196845, 0.688784298719838, 0.364952467614785, 0.313344607595354, 0.911360714351758, 0.604096546536312, 0.566871123155579, 0.639998216181993, 0.648425641935319, 0.341671168338507, 0.107629503356293, 0.729033594718203, 0.373742107534781, 0.981243665562943, 0.177236632211134, 0.207734496565536, 0.442230644868687, 0.391741398023441, 0.667959083337337, 0.764939791290089, 0.723443951457739, 0.427492450457066, 0.565352302277461, 0.280415315181017, 0.216345075517893, 0.472722510341555, 0.849505184916779, 0.958524218061939, 0.527453256072477, 0.263508153613657, 0.0446474615018815, 0.719732520636171, 0.0281985350884497, 0.288600709754974, 0.199770439881831, 0.905538116581738, 0.692372537916526, 0.479486159048975, 0.851432087598369, 0.213101587025449, 0.337060780962929, 0.742256145691499, 0.730832207016647, 0.352350783068687, 0.952726900111884, 0.530918166739866, 0.219672532286495, 0.036447667516768, 0.833013012539595, 0.619245118461549, 0.238139719935134, 0.370675639947876, 0.978851488791406, 0.7006718700286, 0.263727856567129, 0.339404515223578, 0.499661364126951, 0.990128380712122, 0.124457036843523, 0.0241643451154232, 0.489373027114198, 0.332265070872381, 0.0518904554191977, 0.413479605922475, 0.0283480975776911, 0.785610113758594, 0.527596964268014, 0.587601835140958, 0.988093597115949, 0.0310165511909872, 0.560599536402151, 0.140902521554381, 0.415268413955346, 0.755785119021311, 0.530472618760541, 0.88686131965369, 0.889831341803074, 0.0221780149731785, 0.715021695243195, 0.228912900201976, 0.46751338778995, 0.691846843576059, 0.382949794176966, 0.124765643384308, 0.193820160115138, 0.0116696110926569, 0.499008275801316, 0.494003602303565, 0.341777751687914, 0.83416813868098, 0.463684654096141, 0.34608467342332, 0.966182699427009, 0.525997710879892, 0.0441328829620034, 0.530916116200387, 0.976305374410003, 0.989050255157053, 0.0104198292829096, 0.113546064822003, 0.743119424907491, 0.237104511121288, 0.460104257101193, 0.861451941309497, 0.393550509121269, 0.813746106345206, 0.133905489463359, 0.03541862196289, 0.434815890854225, 0.729768058983609, 0.594738359097391, 0.255849457113072, 0.990471224300563, 0.791284440783784, 0.629829506622627, 0.670400384115055, 0.676388737978414, 0.157662501558661, 0.559668902307749, 0.670757708139718, 0.829283415339887, 0.139438835904002, 0.456553161377087, 0.286650779889897, 0.0753583116456866, 0.721121397567913, 0.830062714871019, 0.362020703032613, 0.156580690760165, 0.402641023043543, 0.118217040551826, 0.335816798498854, 0.266962427413091, 0.230960915097967, 0.97327413293533, 0.0901597803458571, 0.44687000173144, 0.745477291755378, 0.135474516544491, 0.102615925250575, 0.104822888737544, 0.748583307256922, 0.212998291477561, 0.245588931255043, 0.801225911825895, 0.225025322055444, 0.489765816368163, 0.643699764506891, 0.396994298323989, 0.327617876464501, 0.382932054344565, 0.77302249148488, 0.583180910674855, 0.493758556200191, 0.745428042951971, 0.238247710745782, 0.579938159324229, 0.34901874512434, 0.434336525853723, 0.542666264344007, 0.0992130602244288, 0.111889486433938, 0.630003085825592, 0.940777980489656, 0.33744064415805, 0.0575709163676947, 0.152528930921108, 0.992703423602507, 0.317453445401043, 0.540752410655841, 0.60876609175466, 0.277596411760896, 0.331576385302469, 0.0795085502322763, 0.718191419262439, 0.784781223861501, 0.88838167139329, 0.86919608223252, 0.289956422289833, 0.191507371142507, 0.222762820310891, 0.3391931084916, 0.135572212515399, 0.37543617724441, 0.438429038506001, 0.572730952873826, 0.989989194786176, 0.950815989635885, 0.929975365288556, 0.277957749553025, 0.710279312450439, 0.385940647916868, 0.115175627870485, 0.22554492438212, 0.077036942820996, 0.942702725762501, 0.959361020475626, 0.587838634848595, 0.166038464521989, 0.395958205452189, 0.987408072222024, 0.498718640534207, 0.022517581935972, 0.0483960495330393, 0.785296971909702, 0.782198448898271, 0.559223517542705, 0.378290116321295, 0.770487060537562, 0.781306929187849), actual = c(1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"