Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.3915810        0.3919932 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.3919932        0.4826577        0.4853598 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.265697839204222, 0.603488277178258, 0.599219968076795, 0.558479259256274, 0.0648339781910181, 0.700513403164223, 0.456296079326421, 0.252573504578322, 0.836469034664333, 0.380599618190899, 0.120202627498657, 0.123162582051009, 0.723050126805902, 0.175053976941854, 0.83119404851459, 0.434640869731084, 0.766929054865614, 0.954466554801911, 0.360382891492918, 0.877630084985867, 0.472106342203915, 0.0590329871047288, 0.222531655337662, 0.230476290220395, 0.352022516308352, 0.18388360994868, 0.244281686842442, 0.356974028516561, 0.107013885630295, 0.866375383688137, 0.25394200719893, 0.367517499253154, 0.740189534379169, 0.448277628049254, 0.287656491156667, 0.395834295777604, 0.756513629108667, 0.259216862963513, 0.150119890458882, 0.830384187633172, 0.834173017879948, 0.0982141811400652, 0.525350325042382, 0.90601348481141, 0.981830351985991, 0.952649442246184, 0.737271517282352, 0.368769516935572, 0.631248068297282, 0.597680957289413, 0.0937896105460823, 0.485720503609627, 0.385646618669853, 0.283064343268052, 0.490408253623173, 0.152582941344008, 0.322763232747093, 0.811334304045886, 0.617330526234582, 0.115778319071978, 0.84802833897993, 0.321650140220299, 0.409654595889151, 0.975384229794145, 0.816054252674803, 0.308307054452598, 0.45743151428178, 0.524965121643618, 0.27009306056425, 0.00275854510255158, 0.130947478814051, 0.549778893822804, 0.584764183498919, 0.377463929122314, 0.675161267165095, 0.413506435928866, 0.514014967018738, 0.227999310242012, 0.386925443541259, 0.304132287623361, 0.600689521990716, 0.7429702389054, 0.789576353039593, 0.32760961051099, 0.989450986729935, 0.722224486060441, 0.690479215700179, 0.324247928103432, 0.136112321633846, 0.990100282011554, 0.984037585789338, 0.790867525851354, 0.649888007668778, 0.594353841617703, 0.604932862799615, 0.527644209330902, 0.750262779882178, 0.968872905243188, 0.604799326509237, 0.433824461651966, 0.737640869338065, 0.87034245277755, 0.707959522726014, 0.839088249951601, 0.382510340074077, 0.765126588754356, 0.238743862370029, 0.561477845301852, 0.672392145497724, 0.889383165398613, 0.332322953967378, 0.477212889119983, 0.33335663122125, 0.533650180790573, 0.425706044537947, 0.569223016034812, 0.0284616672433913, 0.153616743162274, 0.200366608565673, 0.0663260936271399, 0.797287756577134, 0.938988460460678, 0.067105436231941, 0.737696591299027, 0.867779606254771, 0.95096224010922, 0.858004794688895, 0.0370578137226403, 0.775401543593034, 0.770348869729787, 0.460704715922475, 0.140155622037128, 0.128127765376121, 0.384836245561019, 0.948766449932009, 0.880349248414859, 0.74225986097008, 0.394029170041904, 0.829486339353025, 0.456387136830017, 0.228608835255727, 0.744046642212197, 0.477353866444901, 0.759077298687771, 0.824289680458605, 0.889007987454534, 0.691604221472517, 0.0402712314389646, 0.18887230893597, 0.357216814300045, 0.129143057158217, 0.117499015759677, 0.134092058055103, 0.162980142049491, 0.388759109191597, 0.794679000973701, 0.33031651424244, 0.0285724967252463, 0.291959788417444, 0.756918021244928, 0.722577733919024, 0.694708425318822, 0.973289303248748, 0.161203544121236, 0.465938082896173, 0.705330363940448, 0.41190849035047, 0.431487966328859, 0.0692279131617397, 0.55081478622742, 0.419981128536165, 0.978155543096364, 0.592468407237902, 0.692089882446453, 0.965213657356799, 0.0806420645676553, 0.660135878482834, 0.396039119455963, 0.717832532245666, 0.790372218005359, 0.0611369025427848, 0.812956782989204, 0.575658011715859, 0.425377265550196, 0.0358517873100936, 0.927171499002725, 0.048257457325235, 0.517963255988434, 0.783324311254546, 0.912264560116455, 0.934680860722438, 0.902219935320318, 0.345863216090947, 0.0430082061793655, 0.401864947285503, 0.625746140023693, 0.0424722789321095, 0.867418620036915, 0.452950927894562, 0.589278601575643, 0.375339447287843, 0.883539614500478, 0.264515346614644, 0.218242690432817, 0.198543912963942, 0.796124418498948, 0.259365117643028, 0.902193449437618, 0.135236258385703, 0.486182621447369, 0.8168020688463, 0.298393206205219, 0.471344490069896, 0.141212361399084, 0.810582081088796, 0.587550431257114, 0.573202304542065, 0.987555776489899, 0.492571407696232, 0.919329311698675, 0.794116129633039, 0.70787789602764, 0.414263206999749, 0.673163220984861, 0.305311407661065, 0.307136661373079, 0.413367829285562, 0.313627271447331, 0.372429468436167, 0.0411385439801961, 0.646111024776474, 0.647565968567505, 0.146509912097827, 0.528802174609154, 0.44517069729045, 0.653615509625524, 0.598768641240895, 0.362533885985613, 0.91498518595472, 0.25981863681227, 0.236297427210957, 0.113076467998326, 0.424630691297352, 0.141725213034078, 0.329107204684988, 0.683677927823737, 0.103232709690928, 0.263132197549567, 0.964925593696535, 0.231742556672543, 0.592284626094624, 0.452365204459056, 0.557928694179282, 0.733924763277173, 0.482203093823045, 0.574271168326959, 0.901153157232329, 0.764881252776831, 0.946407342096791, 0.504360753577203, 0.102685212157667, 0.329411337617785, 0.755543723236769, 0.972565600881353, 0.0826752923894674, 0.66392902424559, 0.128998944303021, 0.100899196695536, 0.0166125309187919, 0.0234890298452228, 0.886250998359174, 0.442126304609701, 0.700057061156258, 0.830911179073155, 0.352855520090088, 0.813599010463804, 0.0620175185613334, 0.466334197670221, 0.00154888979159296, 0.65684504294768, 0.00477913557551801, 0.174773310311139, 0.620195033960044, 0.00970200286246836, 0.12556512397714, 0.560160772176459, 0.342145098606125, 0.425331006757915, 0.880880078766495, 0.451049134600908, 0.281683567911386, 0.386850890005007, 0.121986811980605, 0.591956531628966, 0.902128684334457, 0.49127357872203, 0.300519892247394, 0.257147618802264, 0.777363269822672, 0.596712801139802, 0.333904586732388, 0.646088391775265, 0.684247850440443, 0.243896098341793, 0.869222035864368, 0.973095889668912, 0.77031455328688, 0.159898520912975, 0.911575669888407, 0.298446264816448, 0.0384110882878304, 0.659409277839586, 0.992355916416273, 0.694451478775591, 0.394488413352519, 0.76502736704424, 0.723639678908512, 0.34656454436481, 0.731693125097081, 0.527484684018418, 0.0705604332033545, 0.413127742707729, 0.697363942628726, 0.0105865844525397, 0.65976919955574, 0.549454777734354, 0.174365003826097, 0.91565823298879, 0.159375179093331, 0.238998324610293, 0.101769466884434, 0.777484715217724, 0.355683093657717, 0.47773189120926, 0.486599070485681, 0.985915009863675, 0.452346275327727, 0.234543633880094, 0.707127180183306, 0.58892990462482, 0.0215554716996849, 0.789485087851062, 0.723586937878281, 0.92979410616681, 0.168373638531193, 0.043720496352762, 0.190011959988624, 0.20715614897199, 0.670011162059382, 0.0757215819321573, 0.310835741693154, 0.521197715774179, 0.424286260735244, 0.129111114889383, 0.00477449153549969, 0.0165244115050882, 0.572203535120934, 0.585465407464653, 0.780945566017181, 0.751835901523009, 0.903959067538381, 0.00729199452325702, 0.87731864489615, 0.95773555804044, 0.18136597960256, 0.763468700228259, 0.0191654334776103, 0.0509946527890861, 0.207371239317581, 0.662088342010975, 0.405443008057773, 0.265990482876077, 0.914907396538183, 0.764674932463095, 0.479465997312218, 0.600768363568932, 0.547513687750325, 0.288767852354795, 0.0685148620977998, 0.85804563947022, 0.237489524064586, 0.614474842092022, 0.486128222662956, 0.594662466319278, 0.459424432367086, 0.951317742466927, 0.886800463311374, 0.505655566463247, 0.145602009026334, 0.219381460221484, 0.119468583026901, 0.963860874297097, 0.905171054648235, 0.741867781849578, 0.815281535033137, 0.264717590995133, 0.936816840898246, 0.48280169069767, 0.706384733784944, 0.323077280074358, 0.673034496139735, 0.00242158793844283, 0.355984359048307, 0.770351860439405, 0.0746957866940647, 0.812154003418982, 0.204122284892946, 0.867150753969327, 0.0719675458967686, 0.717040694085881, 0.719185041030869, 0.350580264814198, 0.646054251352325, 0.106341236270964, 0.786349893081933, 0.475425895070657, 0.130473007215187, 0.205140542704612, 0.700266824336722, 0.79948553792201, 0.836838684976101, 0.0653160270303488, 0.789421712979674, 0.6153355804272, 0.733580901985988, 0.0779843393247575, 0.0851921301800758, 0.826326397247612, 0.742855669464916, 0.0741498032584786, 0.325337257934734, 0.660889677936211, 0.153957346687093, 0.113132525468245, 0.0413221276830882, 0.477765884716064, 0.887900243978947, 0.487411097390577, 0.935692346887663, 0.803595101693645, 0.963045438285917, 0.684714643051848, 0.632739944150671, 0.851682421984151, 0.0433595636859536, 0.617464387323707, 0.736446574563161, 0.413244935683906, 0.157712521729991, 0.980618983274326, 0.486658953363076, 0.361417628126219, 0.487406927160919, 0.555665053427219, 0.223683727905154, 0.997252737171948, 0.27812148979865, 0.264077083906159, 0.786556472303346, 0.460949281929061, 0.353414986981079, 0.0255730049684644, 0.480704226763919, 0.00283736339770257, 0.307681790087372, 0.261282792082056, 0.470556004438549, 0.733795295935124, 0.542033568955958, 0.255478950915858, 0.833873277995735, 0.905331259593368, 0.428834376391023, 0.679659994784743, 0.739974681753665, 0.661021671723574, 0.500733819790184, 0.833505438407883, 0.727308525238186, 0.499660318251699, 0.396609584800899, 0.00841212505474687, 0.544916593469679, 0.454805308952928, 0.371344961691648, 0.95781289273873, 0.198330839630216, 0.805148262530565, 0.808347928803414, 0.938115013996139, 0.798088701209053, 0.734862027689815, 0.592276107985526, 0.322403071913868, 0.952699893619865, 0.457776172785088, 0.611361945047975, 0.488647380145267, 0.0298319007270038, 0.122622237540781, 0.309514888795093, 0.422565345885232, 0.621111450949684, 0.309413507813588, 0.474443562328815, 0.659999654861167, 0.898377743316814, 0.124094701604918, 0.711670875316486, 0.676550062373281, 0.576099448371679, 0.494787452742457, 0.617516786092892, 0.0375858193729073, 0.628366496413946, 0.466903667664155, 0.364985098829493, 0.0778263376560062, 0.37158232787624, 0.864020299632102, 0.640845948597416, 0.668381276540458, 0.619258788879961, 0.2896683299914, 0.152489960426465, 0.797650287859142, 0.0890362905338407, 0.0041243121959269, 0.886595664080232, 0.924066998064518, 0.571582815842703, 0.128786329878494, 0.954199556494132, 0.377767693717033, 0.636691438732669, 0.630137689644471, 0.900029016425833, 0.366177703486755, 0.994938188232481, 0.947936346754432, 0.948832805035636, 0.383929638192058, 0.380293098278344, 0.330449820496142, 0.231542725348845, 0.0194567227736115, 0.27427930617705, 0.750908805057406, 0.906477746553719, 0.0199958302546293, 0.16405340959318, 0.390133651904762, 0.126216075150296, 0.377616342622787, 0.760421029292047, 0.181953859981149, 0.620116666890681, 0.748375761555508, 0.576299900887534, 0.55436194431968, 0.256489686667919, 0.560996240237728, 0.108486899640411, 0.612519861198962, 0.0951086087152362, 0.818054473260418, 0.119075417518616, 0.158282143529505, 0.773879949469119, 0.713230664143339, 0.509196243714541, 0.0913909706287086, 0.276995264692232, 0.850030169123784, 0.314845191780478, 0.26386432047002, 0.773358168080449, 0.912090237019584, 0.0770903686061502, 0.253027173923329, 0.875482976203784, 0.931254533817992, 0.820372907677665, 0.999873132444918, 0.155798221006989, 0.884305393323302, 0.99806375359185, 0.0106423278339207, 0.0365394782274961, 0.131291810423136, 0.822683149017394, 0.066519194515422, 0.769226175267249, 0.774966427125037, 0.975427565164864, 0.912279894342646, 0.698394052684307, 0.00464162812568247, 0.836258855648339, 0.266662223730236, 0.0329704605974257, 0.152548047248274, 0.328134618932381, 0.159637238597497, 0.68496808456257, 0.276214820332825, 0.198358403285965, 0.188845212338492, 0.827951384009793, 0.606853035511449, 0.416515497956425, 0.794623810565099, 0.565838732989505, 0.390522320056334, 0.215056969085708, 0.107500495621935, 0.753745636204258, 0.136381711345166, 0.606483690673485, 0.88016583560966, 0.362308290787041, 0.564382262295112, 0.812806454719976, 0.75450037769042, 0.928044382948428, 0.436000352492556, 0.220610968768597, 0.968035218771547, 0.64683781331405, 0.225832039490342, 0.581319791497663, 0.901230283314362, 0.754728344967589, 0.208641305798665, 0.0584030107129365, 0.898627874907106, 0.783944993978366, 0.626486184541136, 0.529876951826736, 0.319394671823829, 0.559832705650479, 0.881900557782501, 0.355896314838901, 0.581931838067248, 0.743292358703911, 0.354999341303483, 0.991039566928521, 0.108359614154324, 0.26917956280522, 0.488700828049332, 0.983944696839899, 0.805072437506169, 0.496402675751597, 0.673730266513303, 0.291221651015803, 0.127506543649361, 0.675910691730678, 0.960815945873037, 0.823390575125813, 0.206775366328657, 0.172636137343943, 0.425656286301091, 0.899285707157105, 0.139160861959681, 0.685006530256942, 0.663699729833752, 0.968938960926607, 0.0154573528561741, 0.933140760753304, 0.839906478533521, 0.151842617429793, 0.121209256583825, 0.746310549322516, 0.149864863604307, 0.0814567590132356, 0.912731313845143, 0.147338393609971, 0.569795391289517, 0.1276758343447, 0.187609023647383, 0.229987194761634, 0.410169356269762, 0.0774062946438789, 0.351641902467236, 0.0940873096697032, 0.149656970053911, 0.380170687334612, 0.00271559087559581, 0.453367540379986, 0.785390149569139, 0.442002144409344, 0.902395486831665, 0.256823378615081, 0.290427556494251, 0.589019034989178, 0.476666929200292, 0.817426247755066, 0.14011066313833, 0.907178621739149, 0.400879580061883, 0.939616174204275, 0.118327285163105, 0.174571783281863, 0.215820565354079, 0.117410779930651, 0.427543620811775, 0.555209386860952, 0.232223826693371, 0.285119219217449, 0.116794296307489, 0.666223895503208, 0.138413671171293, 0.973876706790179, 0.208609853638336, 0.380195893812925, 0.615011058514938, 0.936332940822467, 0.272129180608317, 0.12874917825684, 0.110932773910463, 0.384965299395844, 0.154772079084069, 0.926561892731115, 0.496955738635734, 0.77332418737933, 0.0876968221273273, 0.899759095162153, 0.520135170547292, 0.381367177236825, 0.659284282010049, 0.391735712764785, 0.839960029581562, 0.618794583249837, 0.129111802438274, 0.000940351281315088, 0.461441236315295, 0.993957289960235, 0.129123734310269, 0.405390404630452, 0.680587347364053, 0.0217846119776368, 0.753422005334869, 0.741686148103327, 0.803122258977965, 0.202828710665926, 0.996097321854904, 0.417714808136225, 0.875760709634051, 0.405074906768277, 0.819792164023966, 0.287520638899878, 0.566421152092516, 0.102765332208946, 0.547456701518968, 0.30963534465991, 0.60830553737469, 0.880801300751045, 0.981963653350249, 0.929642425384372, 0.610557530308142, 0.339071884984151, 0.183797496138141, 0.540824724826962, 0.843415401875973, 0.0998872788622975, 0.139076208230108, 0.727157300105318, 0.517302597640082, 0.612161291763186, 0.545486203860492, 0.775596280116588, 0.122787390602753, 0.910701984073967, 0.500609514303505, 0.0736091667786241, 0.510427691275254, 0.750354969641194, 0.730696512153372, 0.927163713378832, 0.140955631621182, 0.583776583662257, 0.610900418600067, 0.616172946058214, 0.351739744190127, 0.267425412544981, 0.418487178860232, 0.343991686357185, 0.348364460747689, 0.771530651254579, 0.0845785464625806, 0.0618551457300782, 0.297011276008561, 0.919133879942819, 0.926594238029793, 0.11673148162663, 0.0592259976547211, 0.828924945788458, 0.973527293186635, 0.210174445528537, 0.999520535347983, 0.751816045958549, 0.0848460949491709, 0.67257485864684, 0.0847561610862613, 0.119964321376756, 0.830430940259248, 0.965151357930154, 0.457762004807591, 0.220556324580684, 0.855006213765591, 0.444842510856688, 0.9504690321628, 0.429819180164486, 0.55171254160814, 0.679550645872951, 0.526210822863504, 0.986321656499058, 0.393526059109718, 0.591751615982503, 0.460041710641235, 0.412636775989085, 0.775977909564972, 0.345262305578217, 0.924318301491439, 0.763046679552644, 0.151309038512409, 0.779724590014666, 0.029467485845089, 0.526344894431531, 0.696690571960062, 0.461085406132042, 0.489163336576894, 0.949123296188191, 0.816762659233063, 0.58091178513132, 0.649133056169376, 0.265961739700288, 0.983942556660622, 0.0656436299905181, 0.974038327811286, 0.933258654084057, 0.830184422899038, 0.407601963030174, 0.911230504047126, 0.78866953519173, 0.632699467241764, 0.781582857714966, 0.241413415642455, 0.248680142685771, 0.96073441603221, 0.565151978284121, 0.862679480109364, 0.266256584785879, 0.406215665861964, 0.310611265478656, 0.64056847547181, 0.181219980120659, 0.632461993955076, 0.177139800507575, 0.358772698557004, 0.519013992743567, 0.479862625943497, 0.40195633051917, 0.62110891379416, 0.660760918166488, 0.538819922599941, 0.872381330002099, 0.817494162125513, 0.306116164196283, 0.108598773600534, 0.97217862913385, 0.515621317317709, 0.445485431468114, 0.501493134303018, 0.710347777232528, 0.302602813579142, 0.212133994093165, 0.918021908728406, 0.609193208394572, 0.0315820372197777, 0.144888738170266, 0.624979221494868, 0.371228217147291, 0.897738450206816, 0.6329771138262, 0.0330040075350553, 0.446523077553138, 0.613010242814198, 0.921666091540828, 0.378543273312971, 0.982756653334945, 0.667304717935622, 0.924670034786686, 0.855742398882285, 0.159556386526674, 0.820754336426035, 0.612932469230145, 0.325023725163192, 0.100544759072363, 0.892871824558824, 0.454506090842187, 0.238559846766293, 0.448306300677359, 0.961048293393105, 0.204239959362894, 0.769206494558603, 0.194762150058523, 0.542962149018422, 0.805280678905547, 0.922013905132189, 0.220681099686772, 0.334151589311659, 0.166212850715965, 0.855894737876952, 0.0151115220505744, 0.366459843236953, 0.730808113701642, 0.0366497249342501, 0.773788157152012, 0.633658774429932, 0.807208913844079, 0.228339114459231, 0.421987749170512, 0.252400165423751, 0.327410709811375, 0.638011928880587, 0.465619170106947, 0.0619543441571295, 0.284493885701522, 0.71616779291071, 0.967324396595359, 0.263784883543849, 0.533968651667237, 0.102784182876348, 0.5572631850373, 0.503766042878851, 0.396012218669057, 0.701169516891241, 0.74423436797224, 0.946082724491134, 0.505469148978591, 0.279998152283952, 0.0636659723240882, 0.561171482084319, 0.693248629104346, 0.313750475645065, 0.312261675251648, 0.422199960332364, 0.347797549562529, 0.767095170915127, 0.769175234949216, 0.0823409636504948, 0.407188069773838, 0.437826606212184, 0.263044740771875, 0.572651097085327, 0.277969726594165, 0.606540598208085, 0.434035813668743, 0.672341010067612, 0.528475320432335, 0.012369075557217, 0.587545863585547, 0.974200590746477, 0.0914485731627792, 0.339527020696551, 0.0753021682612598, 0.930437793256715, 0.213061606278643, 0.835101350676268, 0.905632532201707, 0.117479497101158, 0.948451075702906, 0.844400281552225, 0.708935600705445, 0.46245691482909, 0.54115127096884, 0.802226624684408, 0.11538154957816, 0.939131410093978, 0.246998498681933, 0.271852447651327, 0.096870951121673, 0.281912362203002, 0.8423983277753, 0.250457277521491, 0.978894956875592, 0.120464434847236, 0.435352347558364, 0.907196967629716, 0.657342153834179, 0.737371923867613, 0.780523418216035, 0.0256618550047278, 0.095592692727223, 0.249324138974771, 0.0534356785938144, 0.315438081044704, 0.54704351676628, 0.473085559671745), actual = c(0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.701146255945787, 0.15687847090885, 0.94636451266706, 0.296714358730242, 0.104361156234518, 0.475272673647851, 0.550230466527864, 0.376570302294567, 0.88984536100179, 0.436740885023028, 0.725657500792295, 0.231049944180995, 0.374792922986671, 0.565670005744323, 0.901891459012404, 0.883625611895695, 0.463106228969991, 0.313823407050222, 0.308120843488723, 0.241200073854998, 0.560489406110719, 0.662638878915459, 0.585225738584995, 0.325818518642336, 0.125046602217481, 0.252775931498036, 0.412948114098981, 0.515324716921896, 0.70137684373185, 0.854265637695789, 0.475257911719382, 0.742054584203288, 0.900181411067024, 0.903110856888816, 0.926835429854691, 0.59054488549009, 0.0112760085612535, 0.0257915570400655, 0.370333707425743, 0.705420569283888, 0.218419482465833, 0.204464635578915, 0.568182791350409, 0.971246971283108, 0.956898023374379, 0.959424076601863, 0.00707679241895676, 0.997561570256948, 0.0918622883036733, 0.316374839982018, 0.499255898641422, 0.400914907222614, 0.0757989012636244, 0.0571960392408073, 0.852170357946306, 0.0791041753254831, 0.874610106693581, 0.125698101241142, 0.51550816046074, 0.972835150081664, 0.652624411741272, 0.549846013775095, 0.688581343274564, 0.434680022997782, 0.262527617393062, 0.124487226130441, 0.414729826617986, 0.316463003167883, 0.0431607882492244, 0.187179648317397, 0.766503702616319, 0.696098136948422, 0.81692774174735, 0.210342517821118, 0.385984561871737, 0.673076989362016, 0.983947156695649, 0.167634167242795, 0.673705232329667, 0.0773896472528577, 0.439297209959477, 0.148239579983056, 0.854943010257557, 0.456687692319974, 0.639345047296956, 0.932794756721705, 0.998970156069845, 0.317984008230269, 0.706473058788106, 0.464548698626459, 0.788292929530144, 0.161791251972318, 0.225784356240183, 0.22086376324296, 0.318818926345557, 0.373078650562093, 0.765635892050341, 0.168013920774683, 0.665905280271545, 0.264739524107426, 0.586144637782127, 0.450746316229925, 0.748704803874716, 0.944874066626653, 0.0444499705918133, 0.62008271063678, 0.0790470934007317, 0.0589684583246708, 0.779383810702711, 0.604325193911791, 0.750170554034412, 0.552404326852411, 0.4131910931319, 0.34881578036584, 0.721889526816085, 0.643383554881439, 0.334605182288215, 0.202157849445939, 0.786418395815417, 0.445965528488159, 0.688854617765173, 0.480706172529608, 0.135070940712467, 0.329311090521514, 0.162542143836617, 0.873163535725325, 0.211624639574438, 0.182786698453128, 0.723657219437882, 0.204739654902369, 0.0562230166979134, 0.142782030859962, 0.821916687535122, 0.764612046768889, 0.830370697192848, 0.535958845401183, 0.304883162258193, 0.747762769227847, 0.582451658556238, 0.553859021281824, 0.161998136434704, 0.935200771084055, 0.492741302121431, 0.113270498579368, 0.629520897055045, 0.311557823559269, 0.376250816276297, 0.187497339444235, 0.743172417161986, 0.433813123963773, 0.080814155517146, 0.715160612249747, 0.262075569014996, 0.321521044941619, 0.571200989885256, 0.0865472606383264, 0.0195776298642159, 0.166944851633161, 0.961658203508705, 0.545078481547534, 0.77434622682631, 0.142591610085219, 0.744281502673402, 0.501791232265532, 0.817279946058989, 0.114586758892983, 0.697973537491634, 0.0694546403829008, 0.62318538594991, 0.649551373207942, 0.789311752421781, 0.899815395940095, 0.74831875343807, 0.906599369365722, 0.926569399656728, 0.472354073310271, 0.289005091879517, 0.659891496412456, 0.323272994253784, 0.109342819312587, 0.330222285352647, 0.751353295752779, 0.89890939113684, 0.407003615051508, 0.662862829631194, 0.355776070617139, 0.831181578338146, 0.809840339701623, 0.455835465807468, 0.804438675986603, 0.654042030451819, 0.643906733719632, 0.864518103422597, 0.465763221029192, 0.76214478071779, 0.803804678842425, 0.662975812796503, 0.701534246327356, 0.909331846982241, 0.482083801180124, 0.937069645617157, 0.291912521235645, 0.585720640374348, 0.229495615931228, 0.0990561838261783, 0.946760471910238, 0.382026027655229, 0.676414806395769, 0.813371000578627, 0.409117808099836, 0.325623069191352, 0.214479382848367, 0.794226801255718, 0.412076865788549, 0.44224381679669, 0.25728168617934, 0.985809676349163, 0.506499595940113, 0.614531652070582, 0.780399012379348, 0.256317695835605, 0.742422106210142, 0.0616892371326685, 0.337796729290858, 0.0375448754057288, 0.251711985794827, 0.0194256072863936, 0.647551181493327, 0.406550824409351, 0.522497542900965, 0.865130266407505, 0.937188026029617, 0.698353645857424, 0.648057103157043, 0.402838780311868, 0.139802067540586, 0.567680547712371, 0.454914169386029, 0.956218604696915, 0.735437369905412, 0.480240453965962, 0.288190752035007, 0.581679214024916, 0.160457787336782, 0.89477961580269, 0.250308303860947, 0.988301980542019, 0.508067275863141, 0.458947267616168, 0.699129270855337, 0.179509360576048, 0.549695389345288, 0.00227976776659489, 0.663565838709474, 0.495736779645085, 0.578482227167115, 0.443042295752093, 0.88886967045255, 0.895315894624218, 0.126726933987811, 0.725029025459662, 0.985838446067646, 0.19029182032682, 0.137839983683079, 0.879092053277418, 0.293302732752636, 0.132536018500105, 0.633088251575828, 0.598858638200909, 0.913988753920421, 0.502860049251467, 0.339121903292835, 0.451552652521059, 0.942512086825445, 0.807595482096076, 0.62816521525383, 0.952642709482461, 0.0559389432892203, 0.3349888632074, 0.773159295553342, 0.815967420348898, 0.292975694872439, 0.0582353938370943, 0.557900839950889, 0.112669684924185, 0.314349734922871, 0.253999510314316, 0.228442759020254, 0.581781129119918, 0.38225734164007, 0.880203601438552, 0.915905819972977, 0.878005279460922, 0.710826665628701, 0.406862745992839, 0.866917395498604, 0.0383447364438325, 0.249679093947634, 0.416767547605559, 0.762193785747513, 0.232611036160961, 0.562397902598605, 0.203676731092855, 0.566163454204798, 0.542516059242189, 0.443421433446929, 0.624397723237053, 0.972919179825112, 0.836264569777995, 0.643685881746933, 0.16904008993879, 0.75361836515367, 0.222325259121135, 0.34156435308978, 0.73420994146727, 0.601771905785426, 0.125453895190731, 0.386267682537436, 0.867770495126024, 0.26489236485213, 0.838199601974338, 0.189927394501865, 0.822708925232291, 0.324860480846837, 0.402749156113714, 0.29326697671786, 0.984437933191657, 0.675386547809467, 0.961691185599193, 0.194750125519931, 0.820833436679095, 0.351762122940272, 0.252651626709849, 0.560685101663694, 0.534551659831777, 0.0433874174486846, 0.988150569144636, 0.932510348036885, 0.485128958942369, 0.56343565043062, 0.488285615574569, 0.200054090935737, 0.975888618733734, 0.320631677983329, 0.705411605536938, 0.304769000271335, 0.0776785081252456, 0.210438936715946, 0.647114270133898, 0.0298421287443489, 0.289230213034898, 0.538257339037955, 0.529098347760737, 0.752169542014599, 0.706258399179205, 0.19379274174571, 0.934774377848953, 0.834622463444248, 0.088868741877377, 0.466446322854608, 0.50842907349579, 0.493903469759971, 0.244483723770827, 0.2814660014119, 0.923579551279545, 0.651139358989894, 0.436700087273493, 0.425547410966828, 0.923953106161207, 0.0909401653334498, 0.562726064352319, 0.113240135833621, 0.145616157446057, 0.386926414445043, 0.248973016394302, 0.0271591546479613, 0.284852660493925, 0.48545120190829, 0.0532603308092803, 0.761027671629563, 0.429346467135474, 0.833551524672657, 0.0327756949700415, 0.354257276514545, 0.287269914057106, 0.146833532489836, 0.740797599777579, 0.958774797152728, 0.769569757627323, 0.738118615467101, 0.482284445082769, 0.578057347564027, 0.0339531376957893, 0.475689109647647, 0.0825564393308014, 0.795672451378778, 0.89188869134523, 0.992617809446529, 0.807543070055544, 0.0150545942597091, 0.412329488899559, 0.218728854553774, 0.241097891470417, 0.74977351911366, 0.586736802943051, 0.574098856421188, 0.986030525295064, 0.793528499547392, 0.65355025860481, 0.996873834868893, 0.08315523667261, 0.250880367821082, 0.797844043001533, 0.400918080471456, 0.0890260296873748, 0.253062459873036, 0.281049834797159, 0.742007975233719, 0.848741912050173, 0.000384006416425109, 0.816912151873112, 0.764451555674896, 0.15289473743178, 0.436423196690157, 0.734598029404879, 0.67123251920566, 0.105763751314953, 0.708361752796918, 0.258905060123652, 0.0219786714296788, 0.00441286899149418, 0.912388552445918, 0.312611514702439, 0.0786382763180882, 0.793369946302846, 0.426844330038875, 0.885995817603543, 0.421407404355705, 0.145174241159111, 0.558362741256133, 0.111104049487039, 0.369073591660708, 0.787684720009565, 0.0779766668565571, 0.224629627075046, 0.448466794332489, 0.739668560214341, 0.360439645359293, 0.58194573642686, 0.690058227162808, 0.542044473113492, 0.991067806724459, 0.358433650340885, 0.606306196888909, 0.116759384982288, 0.919827615376562, 0.305078436154872, 0.551563166547567, 0.706812946125865, 0.650954426964745, 0.350563414162025, 0.491267307195812, 0.8387093055062, 0.328264507930726, 0.357919082744047, 0.923750271322206, 0.776984761003405, 0.0411140504293144, 0.644758491311222, 0.240164143033326, 0.234830648172647, 0.774749162839726, 0.262138781370595, 0.9735391959548, 0.110563294729218, 0.406521417200565, 0.533561459742486, 0.998129977611825, 0.92934409272857, 0.615623062476516, 0.900580454617739, 0.233258807333186, 0.965676666237414, 0.389635482337326, 0.0596748162060976, 0.0692631748970598, 0.672004935564473, 0.194166627014056, 0.0956503469496965, 0.225833277683705, 0.477619926212355, 0.476592983817682, 0.254842049907893, 0.545100275659934, 0.765403403900564, 0.325936308363453, 0.699120832374319, 0.274230248760432, 0.440557996276766, 0.824786063516513, 0.328746145358309, 0.733969953143969, 0.162928561214358, 0.883051766781136, 0.30995174497366, 0.146714980946854, 0.516927164047956, 0.991286187898368, 0.19805292901583, 0.614090361399576, 0.241073890822008, 0.593938796548173, 0.578809153055772, 0.947428134502843, 0.373672806657851, 0.466494211228564, 0.21799877146259, 0.156502897152677, 0.452758198603988, 0.00366950174793601, 0.208670428255573, 0.449189002392814, 0.1173123230692, 0.420781049644575, 0.891914140898734, 0.410976202692837, 0.566670385189354, 0.275607530493289, 0.00933759706094861, 0.909930989844725, 0.515344438375905, 0.228645071620122, 0.557865807088092, 0.0174246125388891, 0.722884759306908, 0.274970370577648, 0.656644843285903, 0.917606646427885, 0.141782574122772, 0.789099931251258, 0.443483086535707, 0.416578539414331, 0.136560304556042, 0.310835536103696, 0.857764103449881, 0.544363273307681, 0.0614047052804381, 0.546800534473732, 0.86572472169064, 0.749346613883972, 0.689160216599703, 0.28680801903829, 0.00751989497803152, 0.729340952355415, 0.536155801033601, 0.889795956201851, 0.36468950053677, 0.196295545669273, 0.964437717106193, 0.382373668253422, 0.209909228142351, 0.194820906035602, 0.153502929257229, 0.48473614314571, 0.583480058470741, 0.339726219885051, 0.317244309000671, 0.437120653688908, 0.454548442037776, 0.47232707682997, 0.362108343048021, 0.253241358092055, 0.416679055895656, 0.906493538524956, 0.298610558267683, 0.165451494976878, 0.791687258053571, 0.566165430704132, 0.640544879483059, 0.141527294414118, 0.775528693106025, 0.15410222299397, 0.68479811004363, 0.0588616107124835, 0.919497567228973, 0.980865701334551, 0.225645698839799, 0.113934664055705, 0.959469458088279, 0.430239452980459, 0.36648198752664, 0.0371283229906112, 0.486426832852885, 0.90159314032644, 0.87465169839561, 0.570734704844654, 0.945153767243028, 0.556790150702, 0.287585733924061, 0.950442276429385, 0.995419539045542, 0.378401794936508, 0.0927247980143875, 0.869272315874696, 0.91881148237735, 0.840962899150327, 0.885948396055028, 0.320964256534353, 0.0898158922791481, 0.398082895437256, 0.61986335623078, 0.101676056860015, 0.627501216251403, 0.108502609888092, 0.691694604465738, 0.235762181226164, 0.645688785705715, 0.0891240055207163, 0.235342331929132, 0.762290521990508, 0.0413034644443542, 0.394902454689145, 0.149725471623242, 0.593080276856199, 0.385287223616615, 0.141187522094697, 0.161974193993956, 0.886130211874843, 0.827058687340468, 0.71668068994768, 0.316054071532562, 0.996683793608099, 0.604024842148647, 0.159302807180211, 0.575829614885151, 0.980144928209484, 0.384160593617707, 0.381603066809475, 0.262609211960807, 0.336065863026306, 0.53473333385773, 0.0797903935890645, 0.911695198621601, 0.816631850553676, 0.0741448199842125, 0.109726463444531, 0.238385034957901, 0.729567256756127, 0.789247593842447, 0.454971921164542, 0.526880328543484, 0.0442951379809529, 0.293638245668262, 0.980912426719442, 0.706757083069533, 0.710810095071793, 0.437414746265858, 0.746324932202697, 0.660410216543823, 0.204709105426446, 0.906081452034414, 0.347245566081256, 0.0674184851814061, 0.108379719778895, 0.279975410783663, 0.553945584222674, 0.788464626995847, 0.842659407528117, 0.95113124535419, 0.349921238841489, 0.477604991290718, 0.0426584107335657, 0.304990568431094, 0.141603532480076, 0.29403696465306, 0.488776893354952, 0.0801528592128307, 0.656381433364004, 0.298618402564898, 0.629697000840679, 0.843967114808038, 0.0453695738688111, 0.708256394602358, 0.145017619477585, 0.0923372630495578, 0.377204184886068, 0.631854766746983, 0.200760128442198, 0.389842974022031, 0.757815046003088, 0.243084935005754, 0.259832862066105, 0.523186338599771, 0.712326742475852, 0.966919013299048, 0.104542490793392, 0.405227427138016, 0.620743753621355, 0.974804637487978, 0.896850548218936, 0.745281491428614, 0.31743828044273, 0.107464603614062, 0.217919517774135, 0.976424493594095, 0.741851361468434, 0.288582912646234, 0.348129286197945, 0.440917707979679, 0.0134504362940788, 0.00600686413235962, 0.565100961597636, 0.991033781087026, 0.489223444834352, 0.58741241786629, 0.522402700968087, 0.237867933465168, 0.35040006483905, 0.297560652950779, 0.0867967160884291, 0.890345648629591, 0.629789893515408, 0.38204019353725, 0.644337157951668, 0.222607517149299, 0.596567099913955, 0.1037041565869, 0.227599578211084, 0.0587661333847791, 0.380479143001139, 0.697074668016285, 0.486291860463098, 0.52889109402895, 0.937687506899238, 0.700400315923616, 0.997637844877318, 0.00101547292433679, 0.504045724868774, 0.334071466932073, 0.952311669941992, 0.628195473458618, 0.550736844073981, 0.707619144348428, 0.464609691174701, 0.379379760008305, 0.232709674863145, 0.50473665096797, 0.413395006442443, 0.75758341117762, 0.477400153875351, 0.0998398596420884, 0.255934123648331, 0.7275311043486, 0.282638606382534, 0.000902986852452159, 0.566439230227843, 0.938255801564083, 0.113622642820701, 0.178263890556991, 0.116795032285154, 0.208979078568518, 0.724488025996834, 0.451993403723463, 0.14515976049006, 0.778930398635566, 0.332549698650837, 0.430786249460652, 0.557738661300391, 0.927742578322068, 0.471635071793571, 0.688366211019456, 0.825769889168441, 0.936353226425126, 0.258127879351377, 0.947235175874084, 0.956475819461048, 0.355585276847705, 0.114458435913548, 0.666546160820872, 0.0721568453591317, 0.612519551068544, 0.683457520790398, 0.966707935091108, 0.778282099403441, 0.394174346933141, 0.314060083124787, 0.904086407506838, 0.0331605507526547, 0.0946542080491781, 0.344970759237185, 0.802613934734836, 0.339223235845566, 0.264588038437068, 0.245733294868842, 0.0562622942961752, 0.621404044795781, 0.945409634150565, 0.883718107827008, 0.913700583623722, 0.27524469117634, 0.820683352882043, 0.774526277324185, 0.214428463019431, 0.457429936388507, 0.543069929350168, 0.936068732989952, 0.968945458065718, 0.412847248837352, 0.869058824842796, 0.458108731545508, 0.546585591277108, 0.0069015440531075, 0.213243550620973, 0.835328072076663, 0.494172015460208, 0.961381802335382, 0.269996943185106, 0.719689751043916, 0.74752227938734, 0.946825724095106, 0.580708026187494, 0.967987030744553, 0.706778598017991, 0.189019044162706, 0.974779026582837, 0.340650079539046, 0.313193402020261, 0.775329569587484, 0.412192481104285, 0.25899846223183, 0.995692236116156, 0.578241900540888, 0.00614765263162553, 0.0904637130443007, 0.450374728767201, 0.729829239659011, 0.460233798716217, 0.690782254096121, 0.63558821612969, 0.868834762601182, 0.900060115149245, 0.879295947495848, 0.429469436174259, 0.354926558444276, 0.316329476889223, 0.707584189251065, 0.714435566216707, 0.0434988448396325, 0.568135153735057, 0.384168164338917, 0.0930988558102399, 0.988068973878399, 0.0468979435972869, 0.171987346373498, 0.695609541377053, 0.433974293293431, 0.658360597677529, 0.345963747007772, 0.940049209631979, 0.140592520823702, 0.642127786297351, 0.100412052823231, 0.563824760727584, 0.596670211758465, 0.752577950479463, 0.0824614923913032, 0.224614091217518, 0.177868046797812, 0.203645886620507, 0.854818622581661, 0.702170669101179, 0.473372625885531, 0.370681886095554, 0.683346559526399, 0.542507428210229, 0.0754014425911009, 0.975802740314975, 0.983548195799813, 0.636972717940807, 0.418004871346056, 0.0569522380828857, 0.0200216437224299, 0.36272252141498, 0.118765487102792, 0.952708896016702, 0.710158081492409, 0.0816362050827593, 0.462686232523993, 0.413823371287435, 0.413253327365965, 0.601696847705171, 0.869403632823378, 0.558672749204561, 0.180392169393599, 0.733474468346685, 0.247036278713495, 0.534648100147024, 0.392363378079608, 0.84337832313031, 0.975429650628939, 0.806499890284613, 0.86577951349318, 0.920331498375162, 0.622360685840249, 0.365501542109996, 0.437465690542012, 0.906755193369463, 0.550718165934086, 0.184087857138366, 0.709587112767622, 0.0301385840866715, 0.663534859893844, 0.785675597842783, 0.722792313899845, 0.146317385602742, 0.838210380170494, 0.21058861608617, 0.751708950847387, 0.739804890705273, 0.50983245135285, 0.377453438472003, 0.871890142792836, 0.354684080928564, 0.201955162920058, 0.785761583130807, 0.880758927902207, 0.394026717636734, 0.762123656924814, 0.227969836909324, 0.631597912870347, 0.861503732390702, 0.391063584713265, 0.709923434304073, 0.607935661682859, 0.164821453625336, 0.811526566976681, 0.938151502516121, 0.197891200194135, 0.174621179001406, 0.515134951565415, 0.896259323228151, 0.0872266888618469, 0.724805025150999, 0.601516545750201, 0.400536995613948, 0.91630282998085, 0.0696656359359622, 0.153580012964085, 0.439017777564004, 0.288748561404645, 0.00360959768295288, 0.955600569024682, 0.696120276814327, 0.3723156650085, 0.388307902263477, 0.633874680381268, 0.297352188499644, 0.867028616601601, 0.981644776416942, 0.329897654708475, 0.324722915887833, 0.635958740022033, 0.109496551565826, 0.146968068322167, 0.644296454265714, 0.666296617360786, 0.692078698659316, 0.849653232842684, 0.106513035716489, 0.655692294007167, 0.412185787223279, 0.278906106250361, 0.948060925351456, 0.676125656813383, 0.985737327486277, 0.923224026337266, 0.984436922706664, 0.104499903740361, 0.404817251488566, 0.0570807564072311, 0.275330121163279, 0.0819031714927405, 0.830900939647108, 0.529426376335323, 0.526774462778121, 0.101894567720592, 0.440672549651936, 0.267798469634727, 0.623982243705541, 0.895474612014368, 0.0344074470922351, 0.0550032355822623, 0.506630928022787, 0.813914143713191, 0.611051640938967, 0.786571825388819, 0.341610561823472, 0.105374206323177, 0.449564249953255, 0.70773639739491), actual = c(1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"