Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4112028        0.4112028 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4897817        0.4897817        0.5170815 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.74502474302426, 0.517037359299138, 0.442750425543636, 0.964900507358834, 0.85487157618627, 0.0998320360668004, 0.56532115698792, 0.463854225585237, 0.629920802311972, 0.102605670457706, 0.607139186235145, 0.161401232006028, 0.907492116559297, 0.0767027197871357, 0.155100213130936, 0.590688653290272, 0.658984625944868, 0.396045428467914, 0.425073244143277, 0.335067030508071, 0.518749268958345, 0.210955346468836, 0.564689155202359, 0.845472735119984, 0.628936138469726, 0.966575132217258, 0.198467449983582, 0.883198074996471, 0.302232207031921, 0.54287137161009, 0.0184866653289646, 0.228831988992169, 0.899896443588659, 0.392902823863551, 0.704946002690122, 0.967766923597082, 0.212635708041489, 0.627101627644151, 0.587155601708218, 0.903032967122272, 0.266134456032887, 0.264942246722057, 0.0771043591666967, 0.295118278125301, 0.0993332122452557, 0.663263431284577, 0.913557389518246, 0.761226628208533, 0.439474873244762, 0.383961333194748, 0.305343553191051, 0.443275251891464, 0.448939275695011, 0.126215974567458, 0.247680983738974, 0.767926146043465, 0.337351223453879, 0.463573214365169, 0.785976398037747, 0.0615640364121646, 0.897229757392779, 0.27029761322774, 0.977871847338974, 0.376854712376371, 0.820401731180027, 0.00111996778286994, 0.697569444775581, 0.0397270733956248, 0.988739133113995, 0.589612814830616, 0.767604388063774, 0.0358100645244122, 0.245379842817783, 0.684516466921195, 0.342102354159579, 0.782649526372552, 0.625989697873592, 0.0908315649721771, 0.668810775503516, 0.642142858821899, 0.64264706755057, 0.190788698848337, 0.103155121672899, 0.712165921460837, 0.521954207681119, 0.880317083792761, 0.580118642421439, 0.726654479745775, 0.699040453881025, 0.240111065329984, 0.0314473055768758, 0.418080237461254, 0.391511847730726, 0.836437088204548, 0.301285423804075, 0.359807181637734, 0.0149769466370344, 0.0567726783920079, 0.392058621160686, 0.500470628961921, 0.452348817605525, 0.766901901457459, 0.469941907096654, 0.630731739103794, 0.835391811095178, 0.473692973377183, 0.245837728492916, 0.11997827119194, 0.868306159740314, 0.277607667027041, 0.436582657974213, 0.998067189706489, 0.595337310805917, 0.658647048054263, 0.72516042762436, 0.559768244624138, 0.911630814429373, 0.860582037596032, 0.928833822952583, 0.207692481344566, 0.211764517938718, 0.796457655029371, 0.361363719217479, 0.179623901378363, 0.113756625680253, 0.134286178043112, 0.811673757620156, 0.393551234621555, 0.897291864268482, 0.55516217299737, 0.988918893272057, 0.307526539312676, 0.207688084803522, 0.0907670306041837, 0.380364270647988, 0.272765658097342, 0.146730022970587, 0.497512870002538, 0.685034541413188, 0.503075575223193, 0.39310031128116, 0.819920454639941, 0.811175059294328, 0.429815224837512, 0.972648029448465, 0.788394016912207, 0.0747552588582039, 0.511011839145795, 0.0696414322592318, 0.79859087546356, 0.361288308864459, 0.111593607580289, 0.657174611464143, 0.872633385006338, 0.686796259367839, 0.633068661438301, 0.738804900785908, 0.856024810578674, 0.151059860829264, 0.916866578627378, 0.670594242634252, 0.317111102398485, 0.248065769439563, 0.174610549816862, 0.488002865575254, 0.627543990267441, 0.394025197718292, 0.0203726899344474, 0.0113485138863325, 0.390458483248949, 0.122792458161712, 0.356377276591957, 0.596943192882463, 0.300494326511398, 0.551499688997865, 0.443538030609488, 0.816264289896935, 0.412423023488373, 0.947493996238336, 0.0660579926334321, 0.463630189886317, 0.993383622961119, 0.350285725435242, 0.076376564335078, 0.0133935180492699, 0.965902753407136, 0.799007158027962, 0.317341628251597, 0.47195131983608, 0.456585839157924, 0.0287784833926708, 0.252956038806587, 0.151598578784615, 0.800128954695538, 0.607135495869443, 0.139106842223555, 0.669845042051747, 0.337279501138255, 0.63696465571411, 0.520834954688326, 0.922976901754737, 0.204894320340827, 0.264121790649369, 0.185248463414609, 0.270123129943386, 0.461071766680107, 0.8397273698356, 0.583388627273962, 0.19879558053799, 0.542931662406772, 0.435071504209191, 0.563493363559246, 0.849890803452581, 0.610761555843055, 0.0697264724876732, 0.45930786523968, 0.196772310184315, 0.709892944898456, 0.889489404391497, 0.298827483318746, 0.612399579491466, 0.0117196363862604, 0.060490996344015, 0.498321343678981, 0.774410976329818, 0.0141002687159926, 0.969219869468361, 0.196202485589311, 0.82916116528213, 0.893320327857509, 0.695265604415908, 0.180798549903557, 0.657687501050532, 0.670233645476401, 0.613036923110485, 0.0402190340682864, 0.597047306830063, 0.70506982691586, 0.0383788773324341, 0.646694025723264, 0.35931160650216, 0.929746241308749, 0.485507228411734, 0.597513587679714, 0.959087959257886, 0.962516001192853, 0.171673306729645, 0.564429289661348, 0.000631526578217745, 0.886746788863093, 0.272083998424932, 0.0252157018985599, 0.639509547967464, 0.970521494746208, 0.903666011057794, 0.867138386005536, 0.784465541597456, 0.975064776372164, 0.239697253331542, 0.538408009102568, 0.899780586129054, 0.398618886945769, 0.755448477342725, 0.758598443586379, 0.972723706159741, 0.456715867388994, 0.463619052898139, 0.0140134275425225, 0.25474087940529, 0.946304977638647, 0.558865900849923, 0.71705134306103, 0.269329310394824, 0.322274044388905, 0.133127511711791, 0.314192956546322, 0.77131526125595, 0.348886679392308, 0.708110751584172, 0.228050135308877, 0.943171679973602, 0.0129204094409943, 0.618473804555833, 0.595652289455757, 0.330222903983667, 0.837268376257271, 0.0186274820007384, 0.365938116563484, 0.83775553223677, 0.542826483258978, 0.854229583172128, 0.435227227862924, 0.837319701910019, 0.0916641948278993, 0.594489623326808, 0.851005150238052, 0.593229480087757, 0.864216389833018, 0.692894847365096, 0.499862891854718, 0.629681317368522, 0.653519286774099, 0.149336739908904, 0.495120965177193, 0.586232817731798, 0.298553285421804, 0.340872750151902, 0.42635117447935, 0.913583679124713, 0.12207102496177, 0.835034838877618, 0.25889925705269, 0.810766395181417, 0.42979293433018, 0.515746284974739, 0.600778030930087, 0.743989329319447, 0.895399128086865, 0.0132178028579801, 0.340621253708377, 0.767342185368761, 0.99665042408742, 0.537183276144788, 0.583500436739996, 0.780105444369838, 0.452034902991727, 0.176392870955169, 0.0201109854970127, 0.374224746832624, 0.2584368721582, 0.369992938125506, 0.961555470479652, 0.569542414043099, 0.521452111657709, 0.0654213794041425, 0.664758781669661, 0.237180729396641, 0.655576853780076, 0.566827374976128, 0.678296522703022, 0.142499414738268, 0.178753477754071, 0.181194630917162, 0.341440109536052, 0.517424604855478, 0.552774820709601, 0.436820867704228, 0.717250044923276, 0.371955351438373, 0.00459319585934281, 0.700060992967337, 0.79332619626075, 0.656386884162202, 0.381739406613633, 0.688548165140674, 0.487708503380418, 0.896752330008894, 0.501279610209167, 0.94595441990532, 0.770526250824332, 0.781554802088067, 0.74605217226781, 0.147015631897375, 0.160128564340994, 0.943087490508333, 0.919371242402121, 0.0772903102915734, 0.907683647470549, 0.496015843935311, 0.349849415244535, 0.321793092880398, 0.17086223536171, 0.533294127788395, 0.0343860010616481, 0.702918989816681, 0.888890715548769, 0.635127308079973, 0.47820083075203, 0.878749787108973, 0.552135444246233, 0.693089979235083, 0.681233602808788, 0.293429676443338, 0.524836074095219, 0.713783725164831, 0.904237869894132, 0.432384227402508, 0.651971150189638, 0.779626992531121, 0.537095986772329, 0.0208030154462904, 0.170451764483005, 0.927338277921081, 0.388608158333227, 0.749531962210312, 0.898456820985302, 0.80764397024177, 0.488347711740062, 0.338737831450999, 0.100822288775817, 0.116998298326507, 0.0260166013613343, 0.99798184633255, 0.154814616311342, 0.803787467535585, 0.342035725712776, 0.9157528441865, 0.821279225870967, 0.915435473900288, 0.183951370418072, 0.877699396340176, 0.342741667060181, 0.33062978903763, 0.899960774695501, 0.211220562690869, 0.0808951151557267, 0.496783212060109, 0.679354786872864, 0.843712050002068, 0.552602040814236, 0.0154698544647545, 0.377330817515031, 0.216990836896002, 0.799032609211281, 0.166389323538169, 0.654258999507874, 0.345543460920453, 0.878935721470043, 0.969509741058573, 0.889741949271411, 0.256053825607523, 0.821826691506431, 0.953203565441072, 0.736033135792241, 0.00520039745606482, 0.631216896930709, 0.0236699031665921, 0.101514565059915, 0.00263700471259654, 0.444413968361914, 0.599924015346915, 0.712338918587193, 0.316923738224432, 0.735653435345739, 0.581373488064855, 0.304195029195398, 0.704393422463909, 0.0263842833228409, 0.942497208947316, 0.749826045008376, 0.883245339617133, 0.56144584948197, 0.156367389718071, 0.256404536543414, 0.223550460767001, 0.497404162772, 0.684862076072022, 0.158603728050366, 0.690444328356534, 0.168128071818501, 0.870103667723015, 0.880670479498804, 0.696878906572238, 0.568526496877894, 0.152408869471401, 0.902301627211273, 0.953497791895643, 0.458154141670093, 0.328642158769071, 0.0857854895293713, 0.568353594280779, 0.498029567534104, 0.270016271620989, 0.836837987415493, 0.801651920424774, 0.427604494150728, 0.0822358964942396, 0.489473835099488, 0.796615711646155, 0.833823976339772, 0.490866971202195, 0.678069189656526, 0.466650738846511, 0.629281317582354, 0.781768881250173, 0.170312017900869, 0.999663496622816, 0.120543340453878, 0.353857851121575, 0.0994777439627796, 0.448575519956648, 0.786285974783823, 0.775802532210946, 0.816317250486463, 0.341013381257653, 0.971864928957075, 0.522379450500011, 0.197228882927448, 0.89751432579942, 0.132303894730285, 0.189959732117131, 0.979361847741529, 0.0325645343400538, 0.81696886010468, 0.640791796147823, 0.343487477395684, 0.972942292690277, 0.432231121929362, 0.623909648042172, 0.309267380973324, 0.275738561525941, 0.810256045078859, 0.70565201016143, 0.942228408763185, 0.425539117306471, 0.419665690744296, 0.0883801206946373, 0.618634383426979, 0.351058603497222, 0.114737986121327, 0.358237331034616, 0.25831168401055, 0.780401236610487, 0.426147860940546, 0.338818168733269, 0.731737992493436, 0.472561501897871, 0.625197391724214, 0.955448522465304, 0.378259152406827, 0.787145690061152, 0.51940030278638, 0.985687069362029, 0.289390076883137, 0.768752818461508, 0.512281552888453, 0.517834559082985, 0.00326843746006489, 0.2359780871775, 0.051218657521531, 0.30477800546214, 0.210307516623288, 0.655826927162707, 0.71824341150932, 0.207934173289686, 0.499162426684052, 0.307458791416138, 0.126352430554107, 0.501414918573573, 0.697423069970682, 0.639845723984763, 0.189026515930891, 0.0196451870724559, 0.630497166421264, 0.0226946934126318, 0.261450105346739, 0.19521226757206, 0.860173407709226, 0.445277692284435, 0.215354459360242, 0.181913659442216, 0.283933807397261, 0.903575651347637, 0.750535209896043, 0.110543388174847, 0.267258684383705, 0.135542346863076, 0.278182549169287, 0.179182366468012, 0.514910871395841, 0.288085646694526, 0.290139273740351, 0.0928114322014153, 0.661131341941655, 0.971336796646938, 0.0863127892371267, 0.487688159802929, 0.900210635503754, 0.809086293447763, 0.164832076523453, 0.265875381650403, 0.548434933647513, 0.116590908030048, 0.830543428193778, 0.952421543188393, 0.00975834438577294, 0.418307591462508, 0.575084640644491, 0.0867488828953356, 0.842320965370163, 0.959354615071788, 0.122855190653354, 0.796188198495656, 0.569976115599275, 0.179167170776054, 0.689009937690571, 0.541688865981996, 0.497758800629526, 0.546102199004963, 0.280992066720501, 0.834421653999016, 0.656053185928613, 0.627020352520049, 0.10509384656325, 0.379588086158037, 0.837072436930612, 0.320507506141439, 0.6637450507842, 0.814388956874609, 0.251733123091981, 0.588734104763716, 0.853020714363083, 0.759737552376464, 0.575981125934049, 0.156568044563755, 0.940184111706913, 0.767232705838978, 0.790166250662878, 0.101946364389732, 0.18636231450364, 0.972605744143948, 0.558606491424143, 0.861706015886739, 0.322603229433298, 0.918948243139312, 0.725495973136276, 0.873212038772181, 0.449253714643419, 0.702853431226686, 0.249410279560834, 0.283858588431031, 0.49568778090179, 0.599015581188723, 0.930188734084368, 0.559481989592314, 0.699922075495124, 0.0850800957996398, 0.260121050756425, 0.136757294414565, 0.669661012245342, 0.928429260849953, 0.774037705501541, 0.598038640106097, 0.0118281892500818, 0.355673850746825, 0.699270417680964, 0.556669707875699, 0.119577712612227, 0.865208526374772, 0.917775081470609, 0.579300069948658, 0.919495463371277, 0.148657317273319, 0.094693282386288, 0.568114094436169, 0.509813528740779, 0.56603591167368, 0.349305438809097, 0.712973845889792, 0.0053160535171628, 0.273636390687898, 0.495899631641805, 0.448841056553647, 0.802861561998725, 0.707988808862865, 0.621282501146197, 0.497268973616883, 0.467308283317834, 0.737116679549217, 0.897812463110313, 0.929066830314696, 0.313164263963699, 0.58444186905399, 0.909685875987634, 0.6633335002698, 0.125640086364001, 0.612712166039273, 0.979543651454151, 0.309430603403598, 0.867055023787543, 0.495391054777429, 0.678498848341405, 0.232340914662927, 0.719506927765906, 0.275823823641986, 0.790734220296144, 0.663640183163807, 0.104472870705649, 0.670787041541189, 0.496854431694373, 0.462423410266638, 0.0151343257166445, 0.594553466653451, 0.611452585784718, 0.452109528239816, 0.462130869971588, 0.740056885406375, 0.606788411270827, 0.303529354743659, 0.428046227199957, 0.856245840666816, 0.146976904943585, 0.0468903386499733, 0.00223823194392025, 0.168949948856607, 0.159259899286553, 0.197647016961128, 0.50977109069936, 0.815650766715407, 0.389498784206808, 0.489840199239552, 0.643628139980137, 0.53524055168964, 0.351676330901682, 0.320978940697387, 0.920423713047057, 0.695695735281333, 0.422838906524703, 0.302556635811925, 0.851176185533404, 0.261646865168586, 0.205504464218393, 0.117526311194524, 0.237573929829523, 0.000380524666979909, 0.925288528902456, 0.487829473800957, 0.246787816984579, 0.832553695188835, 0.159985563019291, 0.145087881013751, 0.217691105557606, 0.383503566728905, 0.0469419346190989, 0.906742698745802, 0.601451376220211, 0.423245104961097, 0.219918935094029, 0.829027800355107, 0.478373684221879, 0.0478509268723428, 0.890241239918396, 0.768958240514621, 0.590836277464405, 0.198082619579509, 0.148662922671065, 0.752915980992839, 0.108276131330058, 0.825620003510267, 0.919305784860626, 0.933237348683178, 0.554098325315863, 0.789252690505236, 0.53596529411152, 0.922548737144098, 0.501696140039712, 0.0899599399417639, 0.588650248944759, 0.961475266609341, 0.186840632930398, 0.496998610906303, 0.113486601039767, 0.665091905510053, 0.513374448055401, 0.741687401896343, 0.734518476761878, 0.527435262454674, 0.967413697391748, 0.344918974908069, 0.0969166294671595, 0.201921871397644, 0.81067630299367, 0.700666291173548, 0.967333956388757, 0.936761676333845, 0.678406038088724, 0.470038384897634, 0.911118236603215, 0.128231227397919, 0.778767693089321, 0.400436932919547, 0.289592451648787, 0.423319909488782, 0.481815577950329, 0.627470418112352, 0.742181897163391, 0.196599721442908, 0.898740606382489, 0.726819657487795, 0.283888766774908, 0.5242996790912, 0.608360145473853, 0.698038283037022, 0.878132657380775, 0.116682403953746, 0.286446423269808, 0.14233578951098, 0.743768718326464, 0.136639013653621, 0.38398932851851, 0.656528992345557, 0.181238556746393, 0.892203410621732, 0.842880406184122, 0.59570339252241, 0.577869126340374, 0.102967990096658, 0.630963862873614, 0.591636280762032, 0.592923437012359, 0.99432921805419, 0.0862827468663454, 0.245723858242854, 0.398172447225079, 0.661047706613317, 0.575759959872812, 0.161183806601912, 0.150563050992787, 0.404393425444141, 0.63309410540387, 0.640734664164484, 0.890112785156816, 0.292680248618126, 0.923426002729684, 0.247639852808788, 0.892301305197179, 0.289706199197099, 0.942112780641764, 0.0674733612686396, 0.366371010197327, 0.618744991719723, 0.866916911443695, 0.402479691663757, 0.466112767811865, 0.720245466800407, 0.182640011887997, 0.0391432871110737, 0.749976058723405, 0.166123606497422, 0.613929544575512, 0.51572615141049, 0.694218208314851, 0.622805105987936, 0.774458604864776, 0.303090506466106, 0.0514759437646717, 0.421888126060367, 0.715517999604344, 0.262212028261274, 0.893265217309818, 0.846369112841785, 0.347317499341443, 0.675128707429394, 0.588289548410103, 0.614079278195277, 0.138682957272977, 0.537303007673472, 0.738455809419975, 0.498415503883734, 0.522576968651265, 0.558230893919244, 0.701727211242542, 0.147658615373075, 0.678283822955564, 0.660759083926678, 0.157339746132493, 0.864355276571587, 0.137653019279242, 0.319557613693178, 0.574895385885611, 0.308352808002383, 0.156273626256734, 0.799951728433371, 0.168455955805257, 0.173569442005828, 0.973667861195281, 0.18157461588271, 0.557796406326815, 0.878497056430206, 0.324746347730979, 0.0167367865797132, 0.808824425097555, 0.734923368319869, 0.929225181229413, 0.603088636649773, 0.535702621098608, 0.779919629450887, 0.831960340263322, 0.268421417102218, 0.322544568451121, 0.486277921125293, 0.7866906367708, 0.206535222940147, 0.621860174462199, 0.0192588437348604, 0.26515780412592, 0.793133009457961, 0.768668077886105, 0.677540875505656, 0.560513076139614, 0.803511995589361, 0.436759323114529, 0.0653466915246099, 0.104675736278296, 0.947266196133569, 0.0918168243952096, 0.405814732890576, 0.0564858289435506, 0.431517806835473, 0.943189488723874, 0.13013968616724, 0.78021400119178, 0.340307120000944, 0.0556202561128885, 0.401154912542552, 0.65208724164404, 0.446662296541035, 0.195776611566544, 0.497784366831183, 0.687752605415881, 0.361730781383812, 0.0882489148061723, 0.821159502258524, 0.186885059578344, 0.320531075354666, 0.641069203149527, 0.180903528816998, 0.836711938958615, 0.547885301988572, 0.165873602731153, 0.0232376845087856, 0.623657379997894, 0.689062625402585, 0.472477046307176, 0.0621650984976441, 0.243452057009563, 0.629091329872608, 0.172828115988523, 0.641580832656473, 0.447255855193362, 0.0266949303913862, 0.58926738309674, 0.166578693548217, 0.853200615849346, 0.926129397936165, 0.0993278087116778, 0.931241324637085, 0.256925809197128, 0.945493474602699, 0.950665762415156, 0.472333605401218, 0.154332045000046, 0.128015003399923, 0.344707380281761, 0.154826791258529, 0.946793541312218, 0.924889424815774, 0.117190711433068, 0.234645873308182, 0.288464258424938, 0.668806981528178, 0.463137059472501, 0.587064958177507, 0.20652267918922, 0.783859381452203, 0.153665371006355, 0.304920652648434, 0.395102974027395, 0.865902466699481, 0.71912138024345, 0.662928320467472, 0.424291156698018, 0.465209370478988, 0.958454229636118, 0.659093115478754, 0.507929380983114, 0.41860327986069, 0.0612263027578592, 0.22565999138169, 0.339880355400965, 0.537642451701686, 0.241873579798266, 0.536384433507919, 0.91135976370424, 0.266417320352048, 0.234746471280232, 0.582376049365848, 0.798711550654843, 0.80086933914572, 0.474944494199008, 0.619622703874484, 0.899658350506797, 0.557468289742246, 0.0697122504934669, 0.107456591911614, 0.0833662040531635, 0.127322426298633, 0.753071004059166, 0.94157115672715, 0.966891040559858), actual = c(0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.808238582452759, 0.0476583286654204, 0.247057299129665, 0.864419417455792, 0.8569109456148, 0.348947856575251, 0.921389078255743, 0.194697950500995, 0.708587764529511, 0.572639848804101, 0.272876171395183, 0.368501124437898, 0.383438259596005, 0.288272248348221, 0.199560382403433, 0.778475398430601, 0.84523243107833, 0.887616328662261, 0.745683583430946, 0.110290234675631, 0.0860659445170313, 0.645484459353611, 0.351398123195395, 0.362006772775203, 0.540369121590629, 0.501330263912678, 0.169072840828449, 0.765554229961708, 0.553525369148701, 0.645516115706414, 0.986865196377039, 0.180254385340959, 0.674192425794899, 0.606178272981197, 0.0417232061736286, 0.799147367710248, 0.938511890592054, 0.556556673720479, 0.62382436171174, 0.864129974041134, 0.476956802653149, 0.0463964440859854, 0.349533223314211, 0.326058763079345, 0.368533747503534, 0.5723784097936, 0.524655394954607, 0.884526226902381, 0.0207271438557655, 0.670307748951018, 0.374836378498003, 0.119217690546066, 0.265528667718172, 0.763205564348027, 0.763179061934352, 0.750075619202107, 0.559546536300331, 0.128292745910585, 0.18000472523272, 0.73916924581863, 0.709134242497385, 0.633446333929896, 0.7854896010831, 0.740262536099181, 0.693538077874109, 0.20010529900901, 0.942694259108976, 0.678708668565378, 0.322601758409292, 0.243270260747522, 0.435853107599542, 0.383336995495483, 0.425142020452768, 0.708450705977157, 0.554416449740529, 0.881992686307058, 0.149041496682912, 0.811616444494575, 0.917916648788378, 0.595201720949262, 0.548065336188301, 0.494821819476783, 0.365208349423483, 0.207918761996552, 0.874186872737482, 0.87413559644483, 0.171053448226303, 0.716686288826168, 0.398105332395062, 0.885542491218075, 0.521118856268004, 0.599912416888401, 0.208399592665955, 0.361826860578731, 0.85927207628265, 0.00669323746114969, 0.0308297199662775, 0.160756764700636, 0.746620390564203, 0.46785409678705, 0.678702945820987, 0.633301031542942, 0.550932633923367, 0.627514798659831, 0.604921041056514, 0.878538886085153, 0.777190851978958, 0.211932060774416, 0.438555252505466, 0.495041216723621, 0.942054926417768, 0.0849218885414302, 0.807612805860117, 0.887324294541031, 0.00393827981315553, 0.482966904295608, 0.86164979590103, 0.509213312296197, 0.892789243953303, 0.843933609314263, 0.877233291277662, 0.752871627220884, 0.763932112371549, 0.00807115877978504, 0.709547341335565, 0.932752124033868, 0.587907278211787, 0.987750936066732, 0.415010701399297, 0.623626177897677, 0.717562654055655, 0.115919596515596, 0.201430650427938, 0.00211706757545471, 0.0808028832543641, 0.248942359350622, 0.932457492686808, 0.922074419911951, 0.606449163286015, 0.0834646157454699, 0.896977501688525, 0.0229475810192525, 0.312585306353867, 0.82582656131126, 0.632621720200405, 0.345602719811723, 0.201880640117452, 0.552780646132305, 0.473054063739255, 0.127497708657756, 0.878953855019063, 0.913458846043795, 0.0758874451275915, 0.932953153038397, 0.377083711791784, 0.434404116123915, 0.699495436856523, 0.491641400381923, 0.484011607244611, 0.466379486722872, 0.240842107916251, 0.811453277710825, 0.656566301826388, 0.466128535801545, 0.0826847599819303, 0.530551674310118, 0.698459077626467, 0.79976892285049, 0.999025502009317, 0.654038870241493, 0.520638336893171, 0.812602108111605, 0.987822932191193, 0.845400572754443, 0.550375922583044, 0.411418386967853, 0.758492850000039, 0.919886339223012, 0.825695149367675, 0.499849677318707, 0.460270408540964, 0.0499848944600672, 0.0876759383827448, 0.808935471111909, 0.893290220759809, 0.0843822744209319, 0.000247773714363575, 0.637439845828339, 0.140065338462591, 0.629996777977794, 0.90225344710052, 0.0936432885937393, 0.693205048097298, 0.215850044274703, 0.352778423111886, 0.267844385234639, 0.171814205823466, 0.203862369293347, 0.630888812243938, 0.0188184373546392, 0.0629603748675436, 0.254245135933161, 0.837678784504533, 0.823522741440684, 0.95928574539721, 0.0682318890467286, 0.610260992078111, 0.64939772314392, 0.531743454979733, 0.704464653274044, 0.371156831271946, 0.355775139294565, 0.79129970446229, 0.424643722828478, 0.0871216505765915, 0.709435776108876, 0.085166335105896, 0.741634269943461, 0.314813168020919, 0.96640502428636, 0.403403245611116, 0.519062384730205, 0.861764355096966, 0.248314913595095, 0.0357010571751744, 0.745558668160811, 0.388408945640549, 0.134677402442321, 0.27462596562691, 0.263360753888264, 0.226246195146814, 0.204947890713811, 0.589920033467934, 0.702154625207186, 0.755006757332012, 0.730243718484417, 0.94415082084015, 0.381872238358483, 0.09629549481906, 0.400104721076787, 0.504828898236156, 0.446948094759136, 0.63763432437554, 0.852724157273769, 0.746698400238529, 0.0123668878804892, 0.410907270619646, 0.601512729423121, 0.127792305545881, 0.85184314684011, 0.171703870408237, 0.805649276124313, 0.207093827659264, 0.272573638241738, 0.328956045908853, 0.221131247235462, 0.480420657433569, 0.470712292939425, 0.746252326061949, 0.412572841858491, 0.378758150152862, 0.413567137904465, 0.681352628627792, 0.874331780942157, 0.247326331445947, 0.956757558509707, 0.16948498156853, 0.883907737443224, 0.403196595143527, 0.433852161979303, 0.111133666476235, 0.879595812642947, 0.813987266737968, 0.832762669073418, 0.141323865624145, 0.513115838868544, 0.751197841251269, 0.319719032384455, 0.92864164756611, 0.245504497550428, 0.205021886620671, 0.00223966664634645, 0.022095883730799, 0.703801804454997, 0.936380531638861, 0.242431388702244, 0.643505381653085, 0.371991379186511, 0.914717776002362, 0.440985353663564, 0.227574483724311, 0.256540726870298, 0.128681732108817, 0.0168399210087955, 0.202831088332459, 0.0969121477100998, 0.192940509645268, 0.724924498004839, 0.672168273478746, 0.580434373579919, 0.57036711089313, 0.487068039597943, 0.208143401658162, 0.992569373454899, 0.846233415417373, 0.912739215418696, 0.85987280565314, 0.765643318882212, 0.000321766827255487, 0.487347051268443, 0.486944978591055, 0.903714532731101, 0.265777210239321, 0.309412555769086, 0.104334849165753, 0.953892324119806, 0.476930679753423, 0.462288391310722, 0.0298207690939307, 0.875743536045775, 0.0221856527496129, 0.441770795499906, 0.810155457351357, 0.880379793932661, 0.676027189940214, 0.348359988303855, 0.351751691894606, 0.109306634403765, 0.728404556866735, 0.444972086930647, 0.679847732419148, 0.244820127496496, 0.655687352642417, 0.716183584183455, 0.345589944394305, 0.656263417331502, 0.259092022897676, 0.681697895284742, 0.37903382605873, 0.625619279919192, 0.426603471161798, 0.772399324458092, 0.263285636901855, 0.410577467409894, 0.515794477192685, 0.604915353469551, 0.58482432179153, 0.319977954495698, 0.378430892014876, 0.381066283676773, 0.373720893403515, 0.097107132198289, 0.345479703973979, 0.814919570460916, 0.60667343554087, 0.957409764640033, 0.277004561387002, 0.862468793755397, 0.733653270406649, 0.866272657644004, 0.274420692585409, 0.371109628817067, 0.7744125793688, 0.96333927102387, 0.409339159727097, 0.922128745121881, 0.813886511372402, 0.864861462265253, 0.818250119220465, 0.422254736535251, 0.810155102051795, 0.44658169336617, 0.98585859593004, 0.199623669032007, 0.186051599914208, 0.571975321508944, 0.675285295117646, 0.808967016404495, 0.0486672292463481, 0.891236184164882, 0.907695135567337, 0.0115325939841568, 0.972290630452335, 0.410835345042869, 0.41361941350624, 0.630970148835331, 0.699038870399818, 0.44193631154485, 0.357057156041265, 0.749372192658484, 0.166150411358103, 0.548502834048122, 0.330018967622891, 0.948843256570399, 0.203468430088833, 0.653405614197254, 0.66401168727316, 0.214752126252279, 0.144529217621312, 0.322998937917873, 0.00981018203310668, 0.173795268405229, 0.83717975136824, 0.562783502740785, 0.210432146210223, 0.0443001247476786, 0.0590801595244557, 0.407630785135552, 0.194849426392466, 0.540599426021799, 0.704952895874158, 0.197600134648383, 0.988676070002839, 0.411114197922871, 0.404791029170156, 0.949438704876229, 0.855953866848722, 0.0845585577189922, 0.761316142044961, 0.865121277747676, 0.471905769547448, 0.114496835740283, 0.520443106535822, 0.293319000396878, 0.313288083998486, 0.693458696827292, 0.915481032337993, 0.835756582440808, 0.833764688344672, 0.603631166508421, 0.0305512282066047, 0.7972439781297, 0.760227057151496, 0.453159558121115, 0.0223793035838753, 0.970228953054175, 0.56809179764241, 0.659597829449922, 0.256549872923642, 0.589764371048659, 0.103366976603866, 0.396596463862807, 0.920123134274036, 0.616476530442014, 0.800757183460519, 0.279126567998901, 0.0240507137496024, 0.0101884356699884, 0.386187238618731, 0.792447526007891, 0.8910298331175, 0.130681187380105, 0.686564141185954, 0.751394228078425, 0.012416142039001, 0.107829330954701, 0.596455921651796, 0.830114493146539, 0.790703132981434, 0.869260743260384, 0.344003545353189, 0.852559885941446, 0.282624270999804, 0.115091542480513, 0.992214780068025, 0.681982991052791, 0.134974814252928, 0.940650292905048, 0.825843372382224, 0.363524788757786, 0.702230155933648, 0.901550189359114, 0.520262291189283, 0.843048599781469, 0.352701028110459, 0.562915624585003, 0.0786208622157574, 0.836352745303884, 0.0401208286639303, 0.975472971564159, 0.691829984774813, 0.165475156158209, 0.209392794175074, 0.113712042337283, 0.0472881568130106, 0.362504043150693, 0.705602357862517, 0.472425502492115, 0.186973902396858, 0.88374161045067, 0.829464942216873, 0.443667753133923, 0.686900391709059, 0.0226599520538002, 0.843308958690614, 0.777998739155009, 0.972442928701639, 0.616439372301102, 0.657426853431389, 0.749820654746145, 0.640275051118806, 0.540140508208424, 0.575760094914585, 0.084660233464092, 0.813652515411377, 0.88755569816567, 0.952273800736293, 0.42227235250175, 0.337267535738647, 0.746954764937982, 0.586517911870033, 0.334684424800798, 0.0654453872703016, 0.343809636309743, 0.314061103388667, 0.610317880054936, 0.633111510192975, 0.249424409121275, 0.232258561765775, 0.7858349038288, 0.759380943141878, 0.67975779203698, 0.167168233310804, 0.296520545845851, 0.334137549856678, 0.492883438710123, 0.00346403126604855, 0.429058163892478, 0.50218834518455, 0.679187265224755, 0.178223235765472, 0.295037460746244, 0.579271643655375, 0.102025342872366, 0.529326780932024, 0.488054873421788, 0.842046553967521, 0.156599968206137, 0.158208774635568, 0.763585457345471, 0.0144757311791182, 0.0792465675622225, 0.268128470284864, 0.221246849745512, 0.594146077288315, 0.879220231203362, 0.644589679315686, 0.996182674309239, 0.560365953017026, 0.312480970285833, 0.434067131951451, 0.665985143510625, 0.645111113088205, 0.239607561845332, 0.0307503915391862, 0.0737990462221205, 0.537139799678698, 0.134857831755653, 0.12176092620939, 0.632664924720302, 0.700752145145088, 0.511547141009942, 0.236105498392135, 0.446710479911417, 0.0515612666495144, 0.842419024556875, 0.899780000327155, 0.186144517734647, 0.994023193605244, 0.159949188120663, 0.30246549542062, 0.960955798393115, 0.126444208435714, 0.692427968839183, 0.330245342338458, 0.251389835728332, 0.748696031747386, 0.398369790520519, 0.519532691687346, 0.765681903110817, 0.370073787402362, 0.366419850615785, 0.411665115971118, 0.99174011638388, 0.52560482500121, 0.523669732967392, 0.653308064909652, 0.116424892330542, 0.496942203957587, 0.897287247003987, 0.451179244788364, 0.843119961908087, 0.143644285388291, 0.683369906386361, 0.815472485963255, 0.863780767656863, 0.220252013299614, 0.721264031948522, 0.412343791220337, 0.936994418269023, 0.510306431446224, 0.279486537678167, 0.00447020190767944, 0.734423791524023, 0.500754169654101, 0.240981554845348, 0.695002788444981, 0.17022866057232, 0.277344255708158, 0.105392422759905, 0.112220641458407, 0.171417289180681, 0.0595703362487257, 0.264893993968144, 0.186235265573487, 0.824007667368278, 0.683060063049197, 0.387828214792535, 0.897258811397478, 0.568967962870374, 0.0604797620326281, 0.104418210452422, 0.485910538351163, 0.752195947570726, 0.49882451328449, 0.23319195327349, 0.122659460175782, 0.409082199679688, 0.348649525549263, 0.124543003272265, 0.842526671942323, 0.855593785410747, 0.83244648296386, 0.174391700187698, 0.221207735128701, 0.518297581933439, 0.522335319081321, 0.357280683005229, 0.732196556637064, 0.899188158335164, 0.254113970091566, 0.0147637303452939, 0.889491340378299, 0.888146370649338, 0.49865424958989, 0.24739292752929, 0.556876898510382, 0.34371007909067, 0.200879975687712, 0.523810344515368, 0.0692551846150309, 0.791837472934276, 0.7767962384969, 0.643357180291787, 0.889599755872041, 0.760352229699492, 0.892925575142726, 0.60421258630231, 0.1445145343896, 0.144595829769969, 0.191047279862687, 0.10719874035567, 0.820374478818849, 0.971299041528255, 0.310416862368584, 0.986538452096283, 0.246376249240711, 0.567339175613597, 0.0849316134117544, 0.466940306592733, 0.280762412352487, 0.898952070157975, 0.659669659333304, 0.699587169801816, 0.638785166665912, 0.739418702665716, 0.152915764367208, 0.458191632060334, 0.0279417338315398, 0.155772493919358, 0.463149734307081, 0.446144199930131, 0.917372986907139, 0.0178998433984816, 0.406047987984493, 0.329721372574568, 0.260585028212517, 0.786732015199959, 0.370586811564863, 0.404292837018147, 0.221659112721682, 0.17449409654364, 0.514799430500716, 0.627308904426172, 0.783278059912845, 0.11895103380084, 0.848768284078687, 0.600891276961192, 0.416285442886874, 0.708647205261514, 0.370460167294368, 0.617609914159402, 0.285202780971304, 0.483307904563844, 0.51266620750539, 0.141242911340669, 0.474619368789718, 0.30101936054416, 0.00555754429660738, 0.306229337817058, 0.132295516785234, 0.185352275846526, 0.628478207858279, 0.208946227794513, 0.147800389910117, 0.136349327163771, 0.0560258401092142, 0.287143533816561, 0.0539431665092707, 0.999573136446998, 0.660834388574585, 0.688531280262396, 0.401998516870663, 0.436974129173905, 0.180284242611378, 0.836014691507444, 0.89596483297646, 0.755065132630989, 0.0673632856924087, 0.88547251955606, 0.884188294643536, 0.811504849465564, 0.771872394485399, 0.0879285759292543, 0.812268501613289, 0.446338678011671, 0.247799065662548, 0.962653690949082, 0.975464126793668, 0.253658674657345, 0.558999527478591, 0.153924038400874, 0.954665305558592, 0.739323226036504, 0.118818966206163, 0.688480153214186, 0.0580826443620026, 0.780618162127212, 0.113706126343459, 0.451692102244124, 0.108187546255067, 0.00257675535976887, 0.754603660665452, 0.0731516776140779, 0.31798959732987, 0.0933189063798636, 0.696293309796602, 0.538014968158677, 0.33091451600194, 0.741063818102702, 0.283814238617197, 0.560088418424129, 0.0734746800735593, 0.976719579659402, 0.656169381923974, 0.355960589367896, 0.574631845345721, 0.490317021729425, 0.89797485456802, 0.991538816830143, 0.0818074909038842, 0.668987466953695, 0.813913030549884, 0.737986418651417, 0.796691167633981, 0.384556993842125, 0.000343670137226582, 0.241527353180572, 0.435155505780131, 0.103941737907007, 0.784481179900467, 0.837265121983364, 0.0961891114711761, 0.822197704110295, 0.444381097797304, 0.857067695586011, 0.894481325522065, 0.757879222976044, 0.0838205346371979, 0.755305707454681, 0.309575012652203, 0.691812435165048, 0.537845635786653, 0.0838531050831079, 0.891618682537228, 0.557067953515798, 0.160226102685556, 0.356197498971596, 0.302885274402797, 0.92003116151318, 0.551874360302463, 0.591123507358134, 0.400522742420435, 0.923001051181927, 0.334512125467882, 0.303228827891871, 0.315404458204284, 0.0748162611853331, 0.296519209397957, 0.727980860741809, 0.42377707734704, 0.734853655099869, 0.62270567077212, 0.904440256534144, 0.955711165675893, 0.798120849998668, 0.170561133651063, 0.0764026048127562, 0.586989180417731, 0.00627477839589119, 0.88626538333483, 0.853967373259366, 0.904209536965936, 0.280747864628211, 0.185709004290402, 0.590719836996868, 0.460892725968733, 0.688690628390759, 0.982603907119483, 0.0673104671295732, 0.21350706089288, 0.52380791073665, 0.477893455186859, 0.153393374290317, 0.579012983711436, 0.483863596105948, 0.249732425669208, 0.841631003189832, 0.00708071631379426, 0.394086835673079, 0.93563246820122, 0.318190373713151, 0.117643068078905, 0.773857820779085, 0.530623098369688, 0.218656658893451, 0.434943622909486, 0.741235533496365, 0.836730571230873, 0.29072216222994, 0.85435997042805, 0.0227239634841681, 0.604516808874905, 0.370278371730819, 0.767989778425545, 0.219402133487165, 0.879156049806625, 0.955176451476291, 0.702703512972221, 0.696225593565032, 0.754230523249134, 0.415137496078387, 0.68326665693894, 0.0427773431874812, 0.749162200372666, 0.64447077549994, 0.362985443789512, 0.656250326894224, 0.479808856500313, 0.912655113032088, 0.0837103435769677, 0.0792371884454042, 0.616839886642992, 0.0949876895174384, 0.699750526575372, 0.00892865238711238, 0.392844602232799, 0.6713728136383, 0.974016229156405, 0.458522849483415, 0.603285513818264, 0.894919587299228, 0.399763049790636, 0.482081843540072, 0.342562151141465, 0.176359050208703, 0.757235953351483, 0.474106436595321, 0.895690734032542, 0.400745423045009, 0.807087829569355, 0.432329056318849, 0.524866377236322, 0.860485650366172, 0.714317485690117, 0.749337198445573, 0.99014885071665, 0.0294184826780111, 0.554736229591072, 0.881723162485287, 0.40910106850788, 0.116971810115501, 0.876895644236356, 0.912228456698358, 0.947248459560797, 0.0232425453141332, 0.878907338017598, 0.290331413270906, 0.773554964689538, 0.59676976245828, 0.195150817045942, 0.625137056224048, 0.0379855383653194, 0.00847770343534648, 0.708414189284667, 0.171602008165792, 0.807907140813768, 0.261706652352586, 0.738312769914046, 0.820976355811581, 0.422088986262679, 0.961063040886074, 0.801049239234999, 0.192428335780278, 0.924563104752451, 0.0318987916689366, 0.82008218485862, 0.033599411835894, 0.3971106284298, 0.78440458746627, 0.161871197633445, 0.195286789210513, 0.424106716178358, 0.854640686651692, 0.305108310421929, 0.419601368252188, 0.536791993072256, 0.527511714491993, 0.399794065859169, 0.381297870771959, 0.672980926930904, 0.594570780638605, 0.644840439083055, 0.327947466634214, 0.110901922453195, 0.494354328373447, 0.394885882735252, 0.850237539969385, 0.809199101524428, 0.996424736920744, 0.289005141472444, 0.862010553712025, 0.135680002393201, 0.230668333359063, 0.32754710293375, 0.346551066497341, 0.354079561540857, 0.988678011111915, 0.678967592073604, 0.341999608790502, 0.291673112194985, 0.331758549902588, 0.16312488168478, 0.175587562844157, 0.640957348281518, 0.37537690019235, 0.14805759023875, 0.310563004342839, 0.855044225696474, 0.61937999422662, 0.346357992617413, 0.5452482374385, 0.617878269404173, 0.313293117797002, 0.798448208952323, 0.22996156080626, 0.629458446288481, 0.382205725647509, 0.00520227150991559, 0.468998499214649, 0.708837675396353, 0.913099351106212, 0.612510157283396, 0.718086553737521, 0.663196746259928, 0.938310055062175, 0.246521639404818, 0.22553429659456, 0.484782005893067, 0.805483551463112, 0.09063316648826, 0.590925338445231, 0.060274769552052, 0.86918601533398, 0.280249408446252, 0.983891150681302, 0.435038021299988, 0.520444736815989, 0.998108389554545, 0.0457493793219328), actual = c(1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"