Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.8557484        0.5736596 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5247351        0.5247351        0.5165188 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.935458126245067, 0.781517937779427, 0.184552545892075, 0.887013116152957, 0.455389171140268, 0.636036606738344, 0.96847233409062, 0.716220266185701, 0.4763255326543, 0.664220007136464, 0.649250727379695, 0.206378720700741, 0.809110132977366, 0.0800274848006666, 0.792783319251612, 0.799493491416797, 0.0726340166293085, 0.0411405952181667, 0.28451509331353, 0.0348450732417405, 0.915281274123117, 0.931796588469297, 0.263894808711484, 0.61246059788391, 0.476055206498131, 0.737337853759527, 0.361511764349416, 0.883117601042613, 0.386724413372576, 0.416662972886115, 0.87552219373174, 0.533639498287812, 0.806737249717116, 0.899502631276846, 0.296282837167382, 0.210563807981089, 0.877812009770423, 0.881494755856693, 0.849649332929403, 0.0789003637619317, 0.931195842335001, 0.60406891698949, 0.421654553618282, 0.506533867446706, 0.0824358779937029, 0.308494001394138, 0.417701966129243, 0.425645706942305, 0.376404466573149, 0.38716184720397, 0.436327189439908, 0.49290247540921, 0.36094310041517, 0.169064463349059, 0.487662418745458, 0.0301258156541735, 0.251565137179568, 0.754228857345879, 0.0844798688776791, 0.892078736331314, 0.187349383719265, 0.0145919371861964, 0.739156230352819, 0.490094005828723, 0.614541521063074, 0.71277197659947, 0.0936426403932273, 0.136742929695174, 0.894598869606853, 0.487451039953157, 0.308683849405497, 0.81979170255363, 0.0428911929484457, 0.18822801602073, 0.722167159197852, 0.0254595798905939, 0.0239337710663676, 0.660172556294128, 0.930913059739396, 0.0614124538842589, 0.312569202156737, 0.847180569311604, 0.959794633323327, 0.967161214444786, 0.75571141182445, 0.167719220509753, 0.469172560377046, 0.315761940786615, 0.440337783889845, 0.664854272268713, 0.717687929980457, 0.47753371600993, 0.254032123833895, 0.755069894948974, 0.159294928889722, 0.428958670469001, 0.105487220920622, 0.150997160002589, 0.0980411635246128, 0.260971260489896, 0.747820440214127, 0.853523155674338, 0.349817734677345, 0.144788508303463, 0.46943099074997, 0.189419982256368, 0.326546973083168, 0.367654170375317, 0.311261138180271, 0.72627515043132, 0.999307334423065, 0.0580229775514454, 0.926009033573791, 0.126418620115146, 0.236369744641706, 0.352157040731981, 0.553151542786509, 0.0234852528665215, 0.886366665596142, 0.616088471142575, 0.819041165988892, 0.678007031790912, 0.761439808644354, 0.717800560407341, 0.605232410365716, 0.749208074761555, 0.125442566582933, 0.455472183180973, 0.5733051833231, 0.569987386465073, 0.667658194201067, 0.299644680693746, 0.511303206905723, 0.992801836691797, 0.14528168621473, 0.798417675774544, 0.209111437667161, 0.641381981549785, 0.860486393794417, 0.641745986882597, 0.993522773496807, 0.807404086925089, 0.480783323990181, 0.289569588145241, 0.495833169901744, 0.835151097970083, 0.330512216314673, 0.872166668763384, 0.773982773767784, 0.838798417011276, 0.870550541440025, 0.081997720291838, 0.53104294044897, 0.455525609431788, 0.173520971089602, 0.585884091444314, 0.70535699557513, 0.353184805018827, 0.590957386419177, 0.554572100285441, 0.517639875877649, 0.582111282972619, 0.773333192570135, 0.297173924744129, 0.552225053310394, 0.243811935419217, 0.497844141209498, 0.244175190804526, 0.758601811248809, 0.59227148653008, 0.804400917841122, 0.454385000281036, 0.699381015030667, 0.491231284104288, 0.879465692909434, 0.350448179990053, 0.264419063925743, 0.704635723959655, 0.789953883038834, 0.94272287003696, 0.421624314039946, 0.933610326843336, 0.0435967384837568, 0.924334259703755, 0.254276041639969, 0.792568944627419, 0.163311056094244, 0.668678447138518, 0.795951812062413, 0.68825759133324, 0.608868401730433, 0.556736940983683, 0.455733995418996, 0.981054231058806, 0.828635696554556, 0.770879745017737, 0.118095701327547, 0.58866212470457, 0.835595198441297, 0.0975731979124248, 0.798514038324356, 0.337781770853326, 0.959857745096087, 0.801147950347513, 0.785155534744263, 0.159031660528854, 0.991105726221576, 0.402802144177258, 0.141273668967187, 0.519045453285798, 0.453225010540336, 0.275854180799797, 0.216655026422814, 0.619237974053249, 0.865693525411189, 0.293553430587053, 0.157754282001406, 0.699471603380516, 0.755740543361753, 0.833836666774005, 0.701311251381412, 0.740095416549593, 0.460844420362264, 0.387943406589329, 0.804192184703425, 0.662056786008179, 0.891943377675489, 0.444554982474074, 0.913397011579946, 0.688831999199465, 0.85667980578728, 0.465370914200321, 0.357550890883431, 0.823906527366489, 0.144576808670536, 0.904262896860018, 0.786922873463482, 0.697260344633833, 0.821253580506891, 0.451820406364277, 0.960086375940591, 0.589790479978547, 0.63486838969402, 0.417141887824982, 0.592959965812042, 0.390436282148585, 0.315373876830563, 0.0813917864579707, 0.693317209603265, 0.200779932783917, 0.205654665594921, 0.111934498418123, 0.325320167001337, 0.023624510737136, 0.811625113012269, 0.582790951710194, 0.413236885797232, 0.697859550593421, 0.500587231013924, 0.681590600172058, 0.48522557830438, 0.936860944144428, 0.187772957840934, 0.657288367860019, 0.605271083768457, 0.55319345369935, 0.612883575959131, 0.53595452546142, 0.117363530676812, 0.309618869097903, 0.929351745173335, 0.830361633561552, 0.491098809055984, 0.198379877954721, 0.304569788044319, 0.700299206422642, 0.402021753136069, 0.00446779909543693, 0.923332192702219, 0.656276871450245, 0.596882198471576, 0.0328304518479854, 0.238690375117585, 0.385132451076061, 0.859053690452129, 0.539321597665548, 0.655515655409545, 0.747221330180764, 0.386327178217471, 0.109838791657239, 0.997429333394393, 0.994969686726108, 0.819351147161797, 0.96090938267298, 0.00678028096444905, 0.484240560326725, 0.225355804199353, 0.0365391965024173, 0.604473705636337, 0.37688747397624, 0.5278160427697, 0.593038039049134, 0.397888461826369, 0.431788247544318, 0.0144869892392308, 0.714920057915151, 0.0487554280553013, 0.101862021256238, 0.748713723383844, 0.312690801918507, 0.214662695769221, 0.534300033701584, 0.773646814050153, 0.604810680262744, 0.645531391259283, 0.801895536715165, 0.372321952600032, 0.252356385812163, 0.209454680327326, 0.139741607941687, 0.325332889333367, 0.584409214090556, 0.05471041216515, 0.306179807987064, 0.480597810819745, 0.920649003703147, 0.106234419625252, 0.147164183901623, 0.690584021620452, 0.127677426440641, 0.134430401958525, 0.111776960780844, 0.680313704069704, 0.939576542936265, 0.552439589984715, 0.936993363546208, 0.266143033280969, 0.545970454113558, 0.0956664378754795, 0.474472026806325, 0.232428960036486, 0.0828580448869616, 0.577197868609801, 0.400147946551442, 0.558066597441211, 0.380210701143369, 0.640307008754462, 0.372673965292051, 0.682583343237638, 0.811428143875673, 0.926663174759597, 0.518320283619687, 0.788328720489517, 0.317200849065557, 0.686338102677837, 0.676710597705096, 0.0357344734948128, 0.352130453567952, 0.113380734110251, 0.764271796913818, 0.324801750248298, 0.928133387118578, 0.482699590967968, 0.343165918253362, 0.206939819734544, 0.821616707835346, 0.58213288616389, 0.133928441209719, 0.976474475581199, 0.55578863970004, 0.545784819638357, 0.756769571220502, 0.485729075036943, 0.977255380479619, 0.556152828037739, 0.384301842190325, 0.462925479514524, 0.892765467986465, 0.592773731797934, 0.456091990927234, 0.472562493057922, 0.197099375771359, 0.669584272895008, 0.416694015264511, 0.591472582425922, 0.892778181703761, 0.784703542478383, 0.310678604524583, 0.718424091115594, 0.236755690304562, 0.546565138269216, 0.638001403771341, 0.649650023318827, 0.495573612861335, 0.331750740529969, 0.159672988113016, 0.114292096812278, 0.756598759675398, 0.711790836881846, 0.207068448420614, 0.444388146745041, 0.666060812072828, 0.347764567937702, 0.855309945298359, 0.642022957792506, 0.916634787572548, 0.510244271252304, 0.521265704184771, 0.0829770385753363, 0.711345646064728, 0.847280728630722, 0.869665232021362, 0.00596677837893367, 0.720055054873228, 0.555765277706087, 0.281802149722353, 0.127007355680689, 0.23034213646315, 0.400920981541276, 0.692253823159263, 0.526492398232222, 0.984872180502862, 0.487491675186902, 0.908744335407391, 0.558377617271617, 0.670993529027328, 0.695187854347751, 0.782983381068334, 0.988389625446871, 0.831050344742835, 0.400962056824937, 0.917568180244416, 0.284448589198291, 0.880248827394098, 0.315554246073589, 0.228924658149481, 0.543332544853911, 0.168084275675938, 0.653625432867557, 0.621105775237083, 0.797174537321553, 0.481411938555539, 0.517277530860156, 0.479341257130727, 0.784046061104164, 0.104234892874956, 0.710097113857046, 0.375180916627869, 0.0529251135885715, 0.277140163583681, 0.595332311466336, 0.184803651412949, 0.376056911423802, 0.59277302143164, 0.29940614849329, 0.398933626944199, 0.237101851264015, 0.0689845089800656, 0.431475708261132, 0.777028199285269, 0.224781244294718, 0.422500788001344, 0.168561824597418, 0.16778419050388, 0.253128137206659, 0.95314341946505, 0.269772621337324, 0.475222873967141, 0.369991008425131, 0.670429817400873, 0.838474585209042, 0.561721773585305, 0.507384559838101, 0.526870044413954, 0.238312010886148, 0.512758260127157, 0.912583314580843, 0.38257582555525, 0.217689609155059, 0.420595852192491, 0.704543059691787, 0.0823323596268892, 0.366153670707718, 0.979753189720213, 0.82864323281683, 0.0978473208379, 0.780064322985709, 0.224328184034675, 0.0261282015126199, 0.832180157070979, 0.123914149357006, 0.696511023212224, 0.264168903697282, 0.29518748098053, 0.482975458959118, 0.605004944372922, 0.542788159800693, 0.540959944715723, 0.498221853515133, 0.362443054793403, 0.971668563084677, 0.581739771878347, 0.465122983558103, 0.0870759498793632, 0.18128952709958, 0.317916794447228, 0.343611175892875, 0.269511404214427, 0.486093850340694, 0.867795544676483, 0.030048742191866, 0.833070372464135, 0.0691547854803503, 0.859907426405698, 0.411289675161242, 0.0965319687966257, 0.539251629030332, 0.396692293928936, 0.566217118641362, 0.251677308697253, 0.913788525387645, 0.278896167408675, 0.791256533702835, 0.64265294582583, 0.626223478000611, 0.449275501770899, 0.891806970117614, 0.387622249312699, 0.545661014504731, 0.52083908719942, 0.332124573178589, 0.126386916730553, 0.560202997177839, 0.640705864876509, 0.819887217832729, 0.756402228260413, 0.389497486874461, 0.773923026863486, 0.732303374446929, 0.497409592149779, 0.549615800846368, 0.979931252077222, 0.900567373493686, 0.256484929705039, 0.367267284542322, 0.285452666925266, 0.315585038391873, 0.258577431319281, 0.525739324046299, 0.644779545953497, 0.616524412063882, 0.420102749951184, 0.175284216413274, 0.774514765013009, 0.414223158033565, 0.958850025199354, 0.305048584472388, 0.874589847866446, 0.552352014230564, 0.857720758300275, 0.864961552899331, 0.636467221658677, 0.116902685258538, 0.896636701654643, 0.903215645579621, 0.991301460657269, 0.147531786700711, 0.955482830060646, 0.908296992769465, 0.363725705072284, 0.819955490762368, 0.8511661530938, 0.869131606305018, 0.083791000302881, 0.110279249027371, 0.610708398511633, 0.179035648703575, 0.648612621705979, 0.150491488864645, 0.605785466730595, 0.0570537440944463, 0.35313460114412, 0.687763144494966, 0.721410293830559, 0.831319980556145, 0.818701308453456, 0.0717755572404712, 0.537564234342426, 0.903377349022776, 0.457574239932001, 0.925944422138855, 0.766351917991415, 0.904130432987586, 0.763360217213631, 0.386858148965985, 0.0348639450967312, 0.896443903911859, 0.517811314668506, 0.460581525694579, 0.771537951892242, 0.209914615610614, 0.899427504045889, 0.901771811535582, 0.00573076633736491, 0.79229680611752, 0.538216312415898, 0.170481529319659, 0.422966130776331, 0.390939734177664, 0.0898998843040317, 0.69799678795971, 0.355736648431048, 0.538664325140417, 0.42869541188702, 0.774798437952995, 0.521432451205328, 0.961494494229555, 0.906246572732925, 0.752115306211635, 0.94407199579291, 0.894551743753254, 0.205149882473052, 0.286828734911978, 0.604581158608198, 0.484558184863999, 0.538289172574878, 0.931946920696646, 0.265637902542949, 0.0874942527152598, 0.905730398371816, 0.325029250001535, 0.572792663006112, 0.366826522862539, 0.803114981623366, 0.429786467691883, 0.516583733726293, 0.209188691806048, 0.0260508335195482, 0.46849249349907, 0.52647710731253, 0.295400070259348, 0.769927995046601, 0.964776309672743, 0.00207412405870855, 0.434370470698923, 0.123820082983002, 0.995661815628409, 0.838151163887233, 0.891938670072705, 0.175616221269593, 0.948629620252177, 0.928944260114804, 0.25401680986397, 0.594906067941338, 0.854543010005727, 0.0897235791198909, 0.901328672654927, 0.133613959420472, 0.0185090661980212, 0.411119434982538, 0.14392671501264, 0.294749176129699, 0.794903112109751, 0.728555588517338, 0.484522908227518, 0.325569176115096, 0.339278075378388, 0.0305247141513973, 0.566144766518846, 0.968140996992588, 0.266565771773458, 0.0567075826693326, 0.706714202417061, 0.00980210606940091, 0.331262936117128, 0.0472197877243161, 0.425235093105584, 0.983369661960751, 0.0651638600975275, 0.253587751882151, 0.807473917957395, 0.424684188095853, 0.283954269485548, 0.425349107244983, 0.371352274203673, 0.45910210465081, 0.0529843864496797, 0.0136375445872545, 0.257565228268504, 0.447039164602757, 0.339863691944629, 0.245430736802518, 0.178774257889017, 0.0487037566490471, 0.14346516225487, 0.886520933592692, 0.26084126601927, 0.00583719508722425, 0.144100812030956, 0.533157450845465, 0.11562813539058, 0.0951100559905171, 0.0324222245253623, 0.0981737093534321, 0.167563015827909, 0.671123536769301, 0.170053958892822, 0.100011782022193, 0.225695515749976, 0.515787580516189, 0.898700513644144, 0.533430768642575, 0.512647568713874, 0.623250568751246, 0.064948377199471, 0.622223096899688, 0.905514686135575, 0.707615538267419, 0.221600921591744, 0.133443486411124, 0.390157229267061, 0.189779258333147, 0.119130359496921, 0.871956387767568, 0.183874662965536, 0.391754030948505, 0.947220112662762, 0.833321361569688, 0.785395654616877, 0.622507000109181, 0.147549260873348, 0.552636598004028, 0.293151370249689, 0.158364671748132, 0.398769030580297, 0.749574617017061, 0.213475588941947, 0.779655518941581, 0.132360079791397, 0.0168330201413482, 0.767631209921092, 0.593590453965589, 0.476646352326497, 0.759017371572554, 0.548662228276953, 0.690465175546706, 0.691738961962983, 0.201711328001693, 0.699529546080157, 0.513149976031855, 0.796497276751325, 0.533827234758064, 0.320961360819638, 0.235781857045367, 0.829724964220077, 0.403411205857992, 0.701187493512407, 0.864563260693103, 0.765737363602966, 0.0724996684584767, 0.15226757642813, 0.414793688105419, 0.31862047710456, 0.949519026326016, 0.481863880064338, 0.425202829763293, 0.195768013130873, 0.543159941211343, 0.542913861107081, 0.916884923819453, 0.708213165868074, 0.687627540901303, 0.473241106141359, 0.0148527615237981, 0.0291559766046703, 0.500279525760561, 0.868985487148166, 0.711564757162705, 0.453368242830038, 0.868103285552934, 0.0273436456918716, 0.421217957278714, 0.360287527553737, 0.633055487880483, 0.930082546547055, 0.660444119712338, 0.481376012321562, 0.951685659354553, 0.287161965155974, 0.43140707956627, 0.132978237001225, 0.355345992138609, 0.177227136213332, 0.769687581807375, 0.562954894732684, 0.621378747513518, 0.648173485416919, 0.687323658028618, 0.758753763046116, 0.26792076905258, 0.477144418517128, 0.151921782176942, 0.00178565294481814, 0.343392247799784, 0.84856338170357, 0.789359838701785, 0.265312345232815, 0.183719274355099, 0.830142670776695, 0.0455709483940154, 0.170014732284471, 0.997153939679265, 0.482227497734129, 0.864436470204964, 0.35591107676737, 0.675783720100299, 0.405796639155596, 0.207462629768997, 0.500831168610603, 0.976270033745095, 0.481616115197539, 0.149507555179298, 0.248357964446768, 0.559469138970599, 0.0782463846262544, 0.0541776160243899, 0.236788297304884, 0.590424574911594, 0.305238099535927, 0.650074675679207, 0.889450244372711, 0.629178776405752, 0.300166821107268, 0.0484460112638772, 0.74022649996914, 0.88938069390133, 0.891391906188801, 0.918564862338826, 0.110213924665004, 0.00239872606471181, 0.604054703377187, 0.794721947750077, 0.0338189825415611, 0.0121137674432248, 0.756286092568189, 0.329120475566015, 0.678520010085776, 0.940116536337882, 0.230063319206238, 0.943163348827511, 0.459166355663911, 0.641905682859942, 0.00746711622923613, 0.934620741056278, 0.0434231290128082, 0.832838540198281, 0.992423490155488, 0.525633038021624, 0.431263125734404, 0.899322651093826, 0.753656382206827, 0.774028193904087, 0.711192107293755, 0.895830694586039, 0.171041537541896, 0.0583449653349817, 0.656269863247871, 0.224879550747573, 0.213109835516661, 0.258089854614809, 0.794265606207773, 0.234196357894689, 0.657003381289542, 0.239926559850574, 0.0984004146885127, 0.605362661881372, 0.486848522210494, 0.00624256045557559, 0.375452459789813, 0.475842405110598, 0.46215043659322, 0.789403577568009, 0.797744524665177, 0.503007020335644, 0.273134019458666, 0.158607687102631, 0.779906228417531, 0.568048178916797, 0.324113534996286, 0.87462963629514, 0.0347210406325758, 0.428089349530637, 0.991968005197123, 0.587499611079693, 0.789997126441449, 0.583509935764596, 0.56764934072271, 0.575505400775, 0.631382732186466, 0.258306941250339, 0.0036857055965811, 0.942611928796396, 0.871359660988674, 0.637041697511449, 0.863719780929387, 0.287947621196508, 0.840055746724829, 0.338923550909385, 0.531502187717706, 0.109373370185494, 0.739034656435251, 0.568284860812128, 0.846043672878295, 0.763648526743054, 0.720854196464643, 0.241196739254519, 0.526695660548285, 0.965187696507201, 0.311706943670288, 0.736982148839161, 0.630047554383054, 0.146618850529194, 0.356459995033219, 0.431639218237251, 0.431487139314413, 0.764364065369591, 0.832359873922542, 0.439148758305237, 0.0327641398180276, 0.00485109444707632, 0.579691278748214, 0.540166262537241, 0.617786103859544, 0.572719838004559, 0.00229098601266742, 0.600292737130076, 0.236185138346627, 0.456679880153388, 0.956327877240255, 0.897704348200932, 0.340442187851295, 0.733586365589872, 0.850030607311055, 0.0121957478113472, 0.503789423266426, 0.00774007383733988, 0.677471387200058, 0.687610246241093, 0.00126816518604755, 0.752744927303866, 0.548947801114991, 0.96932681160979, 0.32696866709739, 0.615467409836128, 0.61553535121493, 0.863742452347651, 0.239983451552689, 0.681993069127202, 0.290323325432837, 0.215378866996616, 0.847351280972362, 0.291238022502512, 0.781570184510201, 0.795632214518264, 0.219062546966597, 0.303245910909027, 0.999718016944826, 0.199876996222883, 0.998483791248873, 0.26535254670307, 0.450531423091888, 0.951027600094676, 0.333046416984871, 0.451986009953544, 0.277617341605946, 0.242214212892577, 0.768653908977285, 0.578193280845881, 0.643544828286394, 0.380546589381993, 0.621424027718604, 0.0676230534445494, 0.119765651412308, 0.591686194064096, 0.956079044612125, 0.726376147475094, 0.2931179322768, 0.0429727367591113, 0.464258599560708, 0.565083721885458, 0.14491376071237, 0.0105127661954612, 0.296285626944155, 0.120909621240571, 0.708272639429197, 0.998169632162899, 0.76846867101267, 0.643995147896931, 0.899016055511311, 0.250745609169826), actual = c(1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.1555777923204, 0.0248105425853282, 0.89925875980407, 0.172309468034655, 0.953949087532237, 0.538307583425194, 0.567219944903627, 0.279384149936959, 0.0795962945558131, 0.559130273060873, 0.281553491950035, 0.474641336128116, 0.405901328427717, 0.162773206131533, 0.168153369799256, 0.280992752872407, 0.544377801008523, 0.0149165655020624, 0.952193338191137, 0.341106590582058, 0.495603845687583, 0.246348981978372, 0.891834882553667, 0.174712629290298, 0.333297269186005, 0.177434847457334, 0.798576193163171, 0.920898960903287, 0.651935685193166, 0.321845470694825, 0.0904229171574116, 0.190492439549416, 0.345867669675499, 0.23090870724991, 0.912220130907372, 0.21985122631304, 0.462211568141356, 0.103525064187124, 0.278324072714895, 0.266410630661994, 0.117785976501182, 0.587955397786573, 0.848883530357853, 0.576297362102196, 0.296167961321771, 0.0221985846292228, 0.678619878133759, 0.135105440625921, 0.713755405973643, 0.222150035900995, 0.120912205893546, 0.520474794320762, 0.258004875155166, 0.909116040915251, 0.324347016168758, 0.203782625729218, 0.789622413692996, 0.404039394808933, 0.735918277641758, 0.0291087552905083, 0.430432100780308, 0.301524302922189, 0.794618802610785, 0.903201991459355, 0.43893350334838, 0.294997939839959, 0.628775082528591, 0.0502895000390708, 0.629876089515164, 0.191860147286206, 0.598519586957991, 0.185461937217042, 0.064149928977713, 0.120491314213723, 0.432396749267355, 0.131299239816144, 0.0040613291785121, 0.17961824988015, 0.393482370767742, 0.121471696998924, 0.624531073030084, 0.690490370849147, 0.198914845706895, 0.661506892880425, 0.224378316896036, 0.266535692615435, 0.970038363244385, 0.10615917481482, 0.548690543277189, 0.919042420573533, 0.580475487280637, 0.826676070224494, 0.235419474542141, 0.832479887641966, 0.459150827024132, 0.141005745157599, 0.740738816326484, 0.578443787293509, 0.258642475353554, 0.110451478045434, 0.230137456441298, 0.766450495459139, 0.491613977588713, 0.537507723318413, 0.183097449596971, 0.469914678251371, 0.0765609585214406, 0.682536478154361, 0.522065311903134, 0.110198192531243, 0.176691093016416, 0.665851908037439, 0.335100360913202, 0.83043572306633, 0.28169647953473, 0.27155697485432, 0.924790789140388, 0.310722388327122, 0.152163798455149, 0.423285820288584, 0.570634486153722, 0.882417964981869, 0.958967289188877, 0.441765474621207, 0.301840575877577, 0.584215799346566, 0.994048338616267, 0.664747772272676, 0.501931966515258, 0.396645153872669, 0.78612719476223, 0.175046449759975, 0.463993202429265, 0.101353115169331, 0.779967883601785, 0.83553255489096, 0.51252822461538, 0.232886497862637, 0.0337178290355951, 0.625628651352599, 0.449224859243259, 0.478943450609222, 0.544088162248954, 0.901054174872115, 0.496267991373315, 0.783405053196475, 0.0240534292533994, 0.242789587005973, 0.900748840766028, 0.536609111819416, 0.782601080602035, 0.14533764636144, 0.247678802115843, 0.872821963392198, 0.293327804189175, 0.646120080957189, 0.704934113426134, 0.859329114900902, 0.451006908668205, 0.487835293402895, 0.976898593595251, 0.443883940577507, 0.707192293135449, 0.6758428204339, 0.87260039197281, 0.386461663991213, 0.554493689443916, 0.338114442070946, 0.622479157755151, 0.537134092999622, 0.887310861609876, 0.661658830707893, 0.364795445464551, 0.200995713705197, 0.270878792041913, 0.0199250031728297, 0.315864771138877, 0.718286625575274, 0.0366573124192655, 0.210356823867187, 0.61953657027334, 0.0274653504602611, 0.86379391909577, 0.583508741576225, 0.590056350920349, 0.939709979575127, 0.685010974528268, 0.912353448336944, 0.273633987177163, 0.0326430655550212, 0.741618354571983, 0.887644196162, 0.401247199857607, 0.136105863377452, 0.77511639916338, 0.0987395031843334, 0.799790114862844, 0.724037770414725, 0.999246654100716, 0.502750373678282, 0.804496834054589, 0.677016787463799, 0.289157562190667, 0.737603632034734, 0.80377052584663, 0.233606583904475, 0.0945200275164098, 0.425881458679214, 0.466010140255094, 0.377354855416343, 0.965923617826775, 0.00401753047481179, 0.762288552243263, 0.206260586157441, 0.136857789941132, 0.779496495379135, 0.771845223614946, 0.170938933035359, 0.471301548648626, 0.312029743101448, 0.735056617064402, 0.258222797187045, 0.167984482366592, 0.700113187544048, 0.35337106953375, 0.593395425006747, 0.210840522544459, 0.927504264051095, 0.235163398552686, 0.640041917562485, 0.118233951507136, 0.413922077510506, 0.171285664429888, 0.246916334144771, 0.533653766615316, 0.246477711712942, 0.407276296289638, 0.692818517331034, 0.273369591683149, 0.0475605484098196, 0.67661455180496, 0.437603131635115, 0.448115347884595, 0.321292474400252, 0.227396567584947, 0.932554631726816, 0.490061703836545, 0.0885221322532743, 0.258002072339877, 0.198968110140413, 0.353077795123681, 0.205138987628743, 0.538238727487624, 0.376731045544147, 0.581024676561356, 0.500289340736344, 0.390917528420687, 0.532081126933917, 0.213061306159943, 0.338016229216009, 0.711083842208609, 0.568739550886676, 0.307552569545805, 0.889786486281082, 0.92174637503922, 0.00872937706299126, 0.203683467349038, 0.427994738332927, 0.538606574060395, 0.68184218253009, 0.117631390690804, 0.955106540350243, 0.158459260128438, 0.803232406498864, 0.298106835922226, 0.830660533392802, 0.24857612955384, 0.00586952897720039, 0.953129106666893, 0.228543419623747, 0.322777530644089, 0.906746161170304, 0.0572058616671711, 0.686213451437652, 0.938629988580942, 0.688827420352027, 0.211614659754559, 0.89444352244027, 0.318520094733685, 0.95804496621713, 0.55411911662668, 0.370947219664231, 0.848090828163549, 0.0127266335766762, 0.483548957621679, 0.816693278960884, 0.169939001556486, 0.00891609815880656, 0.988820357946679, 0.155765085481107, 0.902523235185072, 0.906638787593693, 0.941876992350444, 0.413933873875067, 0.79298927821219, 0.878390714759007, 0.504030670039356, 0.809833053965122, 0.939270628383383, 0.654561012983322, 0.300275310175493, 0.358060204656795, 0.0156862000003457, 0.396127919200808, 0.37950585084036, 0.376699520740658, 0.34807088621892, 0.158921974245459, 0.379434783942997, 0.969687860459089, 0.0652741976082325, 0.714744731783867, 0.378667237702757, 0.656803206773475, 0.853689928306267, 0.0811443268321455, 0.86791743687354, 0.628800025442615, 0.722372997086495, 0.581786124967039, 0.451630322262645, 0.304144122404978, 0.816978222457692, 0.804638494737446, 0.912703797454014, 0.404567596036941, 0.835840780986473, 0.57032739254646, 0.109124106820673, 0.47187566710636, 0.719535395968705, 0.689681813120842, 0.777893877588212, 0.312685469165444, 0.509449571138248, 0.0875863840337843, 0.287859687348828, 0.11627041734755, 0.225647554034367, 0.86830669385381, 0.604679602663964, 0.483326886082068, 0.388430336490273, 0.790248417994007, 0.844561239238828, 0.0384060707874596, 0.275782067561522, 0.178030856419355, 0.582362984539941, 0.131372845266014, 0.672457786044106, 0.135309046367183, 0.734876523492858, 0.348805867834017, 0.329153911909088, 0.483563551679254, 0.796978061785921, 0.111555280862376, 0.831919153686613, 0.0710122517775744, 0.163432599278167, 0.0212562843225896, 0.185763445449993, 0.570344062987715, 0.803819678258151, 0.334245395613834, 0.565796007867903, 0.00269552553072572, 0.709807422244921, 0.119415792636573, 0.0354384493548423, 0.910942389862612, 0.910498078679666, 0.881400150479749, 0.857825914630666, 0.499313044594601, 0.211472429567948, 0.242731857579201, 0.665598202729598, 0.754878595937043, 0.23889522626996, 0.532164922915399, 0.137703296262771, 0.229639338329434, 0.630352028878406, 0.70264348667115, 0.979283086722717, 0.994030757108703, 0.540965297957882, 0.325927982339635, 0.829858398530632, 0.87159120850265, 0.524104596115649, 0.758551059057936, 0.12338315974921, 0.0368600632064044, 0.700704119168222, 0.807312876917422, 0.715126553550363, 0.862429233500734, 0.698471893323585, 0.440303799929097, 0.284682763041928, 0.0619569141417742, 0.350086169783026, 0.624573834473267, 0.949152169749141, 0.10160183836706, 0.35193239338696, 0.0553533025085926, 0.225788595387712, 0.321577201830223, 0.436186132254079, 0.599317521089688, 0.583397926995531, 0.0334841979201883, 0.887048549950123, 0.692483718972653, 0.514751535141841, 0.188119329512119, 0.526776636950672, 0.716318695107475, 0.962557036895305, 0.532712692627683, 0.555261789588258, 0.318174365442246, 0.566837613470852, 0.591751172672957, 0.975103243952617, 0.669165331404656, 0.942859451053664, 0.708452465711161, 0.311673681717366, 0.443961051991209, 0.287983877817169, 0.033239894779399, 0.0419695002492517, 0.140695098787546, 0.976368906209245, 0.826024668989703, 0.130185247864574, 0.878836494870484, 0.536754273343831, 0.620792822446674, 0.903157332446426, 0.0849148447159678, 0.221209455048665, 0.546827143058181, 0.0966616200748831, 0.0825387397781014, 0.834386282600462, 0.0152547312900424, 0.966690653003752, 0.828830844489858, 0.956873764283955, 0.485286285867915, 0.793211934855208, 0.146492233732715, 0.795191094744951, 0.204000320052728, 0.866391464835033, 0.618930385448039, 0.0726980089675635, 0.346779885003343, 0.862109494628385, 0.394869270501658, 0.600022244267166, 0.00800285371951759, 0.535201414953917, 0.864618367049843, 0.299752744846046, 0.0339406183920801, 0.0795331357512623, 0.822344173444435, 0.37334257690236, 0.697618581121787, 0.771500671980903, 0.733453829074278, 0.675105319358408, 0.378710922319442, 0.326629243558273, 0.227076751645654, 0.249044167576358, 0.936171208275482, 0.237300835549831, 0.459932680241764, 0.843168935505673, 0.839737034868449, 0.0278798711951822, 0.387623198563233, 0.000517675653100014, 0.553797444328666, 0.975343259749934, 0.477152684237808, 0.750900133512914, 0.919196392642334, 0.885287784971297, 0.773715315619484, 0.723765918519348, 0.19036387023516, 0.961566940648481, 0.142169030383229, 0.661863491404802, 0.713561777258292, 0.623305910965428, 0.701068280730397, 0.642160619841889, 0.0568841139320284, 0.49473939370364, 0.217686324845999, 0.00286378525197506, 0.314412345411256, 0.878065355122089, 0.0114675683435053, 0.568469594232738, 0.759307634085417, 0.266034097410738, 0.139781777281314, 0.477993948850781, 0.150247778976336, 0.623966022627428, 0.748785696690902, 0.937457247404382, 0.0946677203755826, 0.167332711163908, 0.144398162141442, 0.682109601562843, 0.522887783125043, 0.209882945986465, 0.490057391347364, 0.753518894314766, 0.659171295119449, 0.974006104050204, 0.121809500502422, 0.656375649152324, 0.826592775527388, 0.23277670214884, 0.391893435968086, 0.71046355064027, 0.174022545106709, 0.401937659597024, 0.876719998195767, 0.973265708889812, 0.0610099590849131, 0.176865682238713, 0.910155629971996, 0.936441129771993, 0.998649457003921, 0.443577330093831, 0.846474485704675, 0.00107338395901024, 0.454842775128782, 0.084590126061812, 0.149242840707302, 0.72351113636978, 0.344711228972301, 0.797613905044273, 0.973126675235108, 0.417174948379397, 0.96436339057982, 0.879537121159956, 0.756066236179322, 0.354826787952334, 0.761300362180918, 0.850276665529236, 0.766204573214054, 0.037034458713606, 0.447241354733706, 0.513545641675591, 0.620199265424162, 0.667257433524355, 0.403880505822599, 0.135448976419866, 0.903332140063867, 0.749026708072051, 0.948137803934515, 0.112430792534724, 0.845762232551351, 0.221478878287598, 0.661598878912628, 0.467523672617972, 0.0360285902861506, 0.672635248629376, 0.575003715232015, 0.987660664366558, 0.283897320739925, 0.131492719519883, 0.300520804477856, 0.388482796726748, 0.348957538371906, 0.289724110625684, 0.157098297262564, 0.950556253548712, 0.739656396675855, 0.808668545912951, 0.971716569503769, 0.580179043812677, 0.290662781801075, 0.779941652435809, 0.353194315219298, 0.139792929403484, 0.216762368567288, 0.402919520856813, 0.0972141087986529, 0.0602081250399351, 0.0362816259730607, 0.432290834607556, 0.795966404723004, 0.402556998422369, 0.323429729323834, 0.252841023961082, 0.917468900326639, 0.803320923587307, 0.644638366065919, 0.525906214257702, 0.651973989792168, 0.382458541542292, 0.983653308358043, 0.147708429256454, 0.324456387665123, 0.338453516131267, 0.23041514470242, 0.139876772882417, 0.0605189763009548, 0.726211814442649, 0.659307286608964, 0.882485786918551, 0.33254177798517, 0.790016602026299, 0.382913640933111, 0.697789824567735, 0.330705111846328, 0.565125681459904, 0.677277812501416, 0.459083956666291, 0.311385650420561, 0.409097203984857, 0.508142590289935, 0.948372171958908, 0.270513409283012, 0.785629974678159, 0.509586183587089, 0.0954525079578161, 0.664867813931778, 0.866672947537154, 0.933801285689697, 0.965490261791274, 0.371423449367285, 0.726427621440962, 0.888386933133006, 0.380972260609269, 0.571919814217836, 0.873124654404819, 0.277762143407017, 0.275995763717219, 0.912544354563579, 0.176572620868683, 0.174157995032147, 0.221373659325764, 0.304984347196296, 0.255113932071254, 0.940573542611673, 0.946571222273633, 0.635496518574655, 0.427457117009908, 0.0156515983399004, 0.770682812202722, 0.0557304215617478, 0.264300524257123, 0.0199361068662256, 0.18942817277275, 0.042587541975081, 0.541656816843897, 0.594403045251966, 0.778455420630053, 0.143360503716394, 0.973431102233008, 0.121222457615659, 0.61572540178895, 0.961335069267079, 0.19632446509786, 0.759860006393865, 0.222959288861603, 0.800594306783751, 0.369945929851383, 0.254776520421728, 0.23008097265847, 0.58676107483916, 0.878696325467899, 0.175448233494535, 0.575725361704826, 0.786162883741781, 0.869783428497612, 0.0534228493925184, 0.108485981123522, 0.271856704261154, 0.899690269026905, 0.703684119973332, 0.40126212942414, 0.0438544212374836, 0.608981539262459, 0.730237301439047, 0.0550947950687259, 0.189443280454725, 0.631895546568558, 0.556158801773563, 0.17599109467119, 0.696335948305205, 0.665571170160547, 0.10881295776926, 0.857892226893455, 0.750926155596972, 0.996061450336128, 0.138253033393994, 0.577633953187615, 0.152987797744572, 0.528016284340993, 0.97024055221118, 0.45698100887239, 0.0243366966024041, 0.351778940297663, 0.800022735260427, 0.49932908709161, 0.518257919931784, 0.607791198417544, 0.272341337520629, 0.878248841734603, 0.166261818259954, 0.0207823512610048, 0.173937337007374, 0.974830089369789, 0.774083236930892, 0.132444280665368, 0.534467879449949, 0.274842334445566, 0.769218250177801, 0.851451600203291, 0.59142115758732, 0.62425818038173, 0.644545543007553, 0.715990871423855, 0.93112545250915, 0.958217172883451, 0.329288696404546, 0.163098084740341, 0.287130164448172, 0.00108263222500682, 0.363345025805756, 0.480706102913246, 0.565299249021336, 0.373451371211559, 0.795864858198911, 0.830819576513022, 0.390600834274665, 0.370491118868813, 0.0411791442893445, 0.505848387489095, 0.197557171806693, 0.710014855489135, 0.572335420642048, 0.992502950597554, 0.530632719397545, 0.683637437177822, 0.566750648897141, 0.813392902957276, 0.022055592155084, 0.608210560400039, 0.0977941770106554, 0.465044984361157, 0.905297890305519, 0.268338093534112, 0.152732005342841, 0.284943432547152, 0.136359072988853, 0.796331818215549, 0.963123165536672, 0.300216691335663, 0.592754907673225, 0.407270836643875, 0.22461383533664, 0.327769057126716, 0.91507540922612, 0.465574896661565, 0.470970361959189, 0.763282928615808, 0.192848609294742, 0.930750604486093, 0.420901272911578, 0.24388325936161, 0.541196963051334, 0.291007588384673, 0.599154309136793, 0.268969389377162, 0.481502394657582, 0.432450504042208, 0.417441320139915, 0.3109531479422, 0.314237442566082, 0.577292571542785, 0.0696909897960722, 0.670254665659741, 0.622832505498081, 0.543636765796691, 0.183746180031449, 0.497292550513521, 0.162868191488087, 0.780001600738615, 0.587474240688607, 0.0255112654995173, 0.569784698542207, 0.114855885040015, 0.036568769486621, 0.729631538735703, 0.661784801864997, 0.0733037092722952, 0.728110609343275, 0.164327742299065, 0.449257810367271, 0.440830356441438, 0.983268982963637, 0.987800790229812, 0.388600037666038, 0.361997735220939, 0.508621973916888, 0.44352068984881, 0.425345507450402, 0.133791101630777, 0.976575487526134, 0.988539978396147, 0.926865338115022, 0.135820126160979, 0.469966895179823, 0.23574661812745, 0.352465542964637, 0.432595791993663, 0.911571827484295, 0.637188114691526, 0.0466867864597589, 0.966483676340431, 0.969130145618692, 0.0647342891898006, 0.0910829184576869, 0.42153258738108, 0.141372956801206, 0.0264106804970652, 0.0611788402311504, 0.520187010522932, 0.792865334544331, 0.581896850839257, 0.902336339000612, 0.0284863212145865, 0.786505847005174, 0.859861803939566, 0.384068777086213, 0.536253499798477, 0.407330934423953, 0.358551304787397, 0.655192225938663, 0.448203205363825, 0.00753295794129372, 0.916627078782767, 0.569087788462639, 0.710661763325334, 0.618958719074726, 0.988887766143307, 0.189318905118853, 0.73998808558099, 0.750923387240618, 0.141824652906507, 0.751050736987963, 0.497504265746102, 0.0432573917787522, 0.830012825783342, 0.0458739318419248, 0.265545970993116, 0.778729323996231, 0.430635889293626, 0.816215548431501, 0.832874536514282, 0.411400547018275, 0.653416842222214, 0.542809279868379, 0.251037454465404, 0.247379316482693, 0.490375401219353, 0.108884625602514, 0.717866841237992, 0.000506757060065866, 0.13091322267428, 0.69243226153776, 0.679956238251179, 0.945414948044345, 0.674937455914915, 0.450403539929539, 0.542678784346208, 0.81563940574415, 0.97084865742363, 0.179757477948442, 0.650978953111917, 0.371910567395389, 0.996312648290768, 0.584737080149353, 0.634103436022997, 0.877811298472807, 0.259545401204377, 0.209635196486488, 0.0832744156941772, 0.523987128166482, 0.283279422903433, 0.125883608590811, 0.699142273748294, 0.497145819012076, 0.238165737362579, 0.728128329850733, 0.322365600150079, 0.0173166168387979, 0.00779712642543018, 0.47180339647457, 0.536346864420921, 0.982168380636722, 0.726212655194104, 0.0201491029001772, 0.516631894279271, 0.57417715061456, 0.744766644667834, 0.865637004608288, 0.757927739759907, 0.279387533199042, 0.604902111692354, 0.418946790043265, 0.832748754648492, 0.0132149150595069, 0.560883578378707, 0.595465972321108, 0.106822186149657, 0.622283494332805, 0.266808298649266, 0.00760661624372005, 0.0859603555873036, 0.535105418646708, 0.179421594599262, 0.804869539104402, 0.848282545339316, 0.781468303408474, 0.303261714754626, 0.986937697511166, 0.49851262383163, 0.177116261329502, 0.192926577292383, 0.724670676514506, 0.694375643273816, 0.96586699062027, 0.776447060285136, 0.538021054118872, 0.972891220357269, 0.636062589474022, 0.480124281719327, 0.285289252642542, 0.855886399047449, 0.763386311242357, 0.10044968733564, 0.998204267350957, 0.937796277925372, 0.279933378798887, 0.929206733824685, 0.386318976990879, 0.455794726498425, 0.320429697399959, 0.441152988001704, 0.0244790327269584, 0.962227356387302, 0.497377753490582, 0.797112841159105, 0.256573629332706, 0.729760324116796, 0.640266254078597, 0.103860723553225, 0.683283041231334, 0.562642927048728, 0.91917252773419, 0.457169843837619, 0.816820504609495, 0.50869701593183, 0.971310238819569, 0.64998989738524, 0.73710176628083, 0.454192239558324, 0.672460122266784, 0.899538117460907, 0.0287873791530728), actual = c(0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"