Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.8912588        0.8252418 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.6053527        0.5569822        0.4897400 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.711613802006468, 0.491561292437837, 0.0518135465681553, 0.962819554843009, 0.339136194670573, 0.133873402606696, 0.130180761916563, 0.211325349984691, 0.391355761792511, 0.86259011994116, 0.10253381752409, 0.680562969530001, 0.483168070670217, 0.135995726333931, 0.807119246805087, 0.838524106191471, 0.43403078825213, 0.658692613476887, 0.794193542562425, 0.673832024913281, 0.155520864296705, 0.616805313387886, 0.907897245371714, 0.180047703906894, 0.807188884122297, 0.791777123929933, 0.605579225113615, 0.257616942515597, 0.822364353574812, 0.203953678254038, 0.510325677227229, 0.548305511940271, 0.68609045795165, 0.838837170507759, 0.228417372098193, 0.838728624396026, 0.00712790712714195, 0.436290405457839, 0.662264745216817, 0.0315593658015132, 0.790633929893374, 0.79155357577838, 0.35393723170273, 0.268936686683446, 0.302760783350095, 0.754877376370132, 0.953134233364835, 0.519222041592002, 0.791066588833928, 0.0719352250453085, 0.26037819404155, 0.57213731133379, 0.219457568600774, 0.222438207827508, 0.977392286993563, 0.164201046107337, 0.593556668609381, 0.818671823712066, 0.727934744441882, 0.143002583179623, 0.832677344791591, 0.369209322147071, 0.0526310936547816, 0.020061542512849, 0.341172466753051, 0.878867650404572, 0.938682420412078, 0.199884619098157, 0.844698070548475, 0.404912853147835, 0.430714651010931, 0.9183966726996, 0.861728260992095, 0.121849452378228, 0.177647385979071, 0.123302069725469, 0.0703431251458824, 0.874233166221529, 0.181931598810479, 0.872134129051119, 0.247292538406327, 0.299938793294132, 0.868995147291571, 0.653868338093162, 0.55524942651391, 0.546319009037688, 0.502598190214485, 0.505046437494457, 0.725482777226716, 0.730176775949076, 0.451973944436759, 0.0460933167487383, 0.455208761617541, 0.428580957930535, 0.0239338018000126, 0.417300001950935, 0.655853274976835, 0.519846920855343, 0.649166582385078, 0.447368513327092, 0.602198283420876, 0.290788606507704, 0.356004444183782, 0.199740757234395, 0.480381975648925, 0.967545811785385, 0.134641270618886, 0.815703082829714, 0.734807001659647, 0.516357216052711, 0.211368386866525, 0.826677974546328, 0.320326778804883, 0.240414377534762, 0.948829209199175, 0.238899591146037, 0.00984500255435705, 0.480371118988842, 0.256102610845119, 0.959748913068324, 0.656469155102968, 0.730045922566205, 0.356083306018263, 0.334153821924701, 0.582525482401252, 0.594326824648306, 0.297254695324227, 0.932044571032748, 0.643546901876107, 0.166101003764197, 0.589256695006043, 0.237076065735891, 0.823068698402494, 0.704075440764427, 0.96400277223438, 0.9078662651591, 0.442904426716268, 0.27935103001073, 0.863215421559289, 0.36650656699203, 0.139136939775199, 0.287160244071856, 0.880866711493582, 0.503053688909858, 0.708840158535168, 0.306703086243942, 0.893374548759311, 0.745408127317205, 0.11593781854026, 0.0101840384304523, 0.843795619206503, 0.316081790486351, 0.823409638367593, 0.428129599895328, 0.886242975480855, 0.975487571908161, 0.231862565269694, 0.626774356234819, 0.0708689829334617, 0.450402296381071, 0.786739337258041, 0.499867855804041, 0.238616455812007, 0.83235222636722, 0.700810197740793, 0.324586815200746, 0.760214736452326, 0.292362759821117, 0.784335868898779, 0.760639348765835, 0.227900147903711, 0.991152592934668, 0.0397994690574706, 0.0495217873249203, 0.381207762286067, 0.314816718688235, 0.627491632243618, 0.987243169918656, 0.218960259109735, 0.15386839909479, 0.424256256083027, 0.101869985461235, 0.812125249998644, 0.471995083149523, 0.702877133153379, 0.825203269021586, 0.41854363726452, 0.840479693608359, 0.354787425603718, 0.242515147430822, 0.33935492974706, 0.0898333173245192, 0.678858853876591, 0.544448792003095, 0.850336572155356, 0.587399572134018, 0.327751748031005, 0.267184064956382, 0.920991933671758, 0.470250103855506, 0.761495489394292, 0.684536117361858, 0.789868989959359, 0.22704696864821, 0.161047384608537, 0.737248943420127, 0.779020694317296, 0.634313094196841, 0.423771481728181, 0.536707118386403, 0.85829561133869, 0.62448964593932, 0.765607370063663, 0.074286482995376, 0.549409455852583, 0.367740296060219, 0.916518265847117, 0.257341356715187, 0.48791939439252, 0.147779963444918, 0.893140376079828, 0.00797453103587031, 0.68525749607943, 0.361302175093442, 0.638540738262236, 0.700418697204441, 0.283994404599071, 0.305277102859691, 0.522593428147957, 0.0156473354436457, 0.355190432397649, 0.899570565205067, 0.447361632017419, 0.977230015210807, 0.46791878156364, 0.294968459755182, 0.952188384952024, 0.655582614941522, 0.343746666330844, 0.122021490009502, 0.447088211541995, 0.970882993191481, 0.266926349839196, 0.879066845169291, 0.247533388435841, 0.36073178704828, 0.431559444172308, 0.588116573402658, 0.0591866360045969, 0.509943525772542, 0.456377442227677, 0.00998288788832724, 0.0180176424328238, 0.230804577702656, 0.749692398589104, 0.721751370467246, 0.0544390671420842, 0.135043344926089, 0.197930597700179, 0.245167991612107, 0.222945824731141, 0.980531389825046, 0.0853855151217431, 0.491285416297615, 0.84115314274095, 0.755900950869545, 0.509088325547054, 0.844155369792134, 0.0476788792293519, 0.522033001063392, 0.356294346507639, 0.163235643645748, 0.47194562992081, 0.106504918541759, 0.557218986097723, 0.963012247812003, 0.568642238620669, 0.653196089202538, 0.897091824794188, 0.187464287504554, 0.354713154723868, 0.0651540607213974, 0.429583457298577, 0.770203360589221, 0.417373465374112, 0.599518615985289, 0.844365958124399, 0.867607839638367, 0.284729566192254, 0.54957219818607, 0.279883612412959, 0.20693610329181, 0.125789878657088, 0.757178999949247, 0.646312630502507, 0.250248091993853, 0.984925841912627, 0.595640301471576, 0.657817741623148, 0.431943176314235, 0.691424437332898, 0.790457563474774, 0.522189448587596, 0.311465735081583, 0.411102488869801, 0.577057396061718, 0.667072475189343, 0.692956484621391, 0.449732198845595, 0.061550973681733, 0.929948591860011, 0.256506512174383, 0.623577396851033, 0.250061040278524, 0.189307955093682, 0.401053868699819, 0.386869277106598, 0.433249853551388, 0.935005000093952, 0.75919125508517, 0.107257130090147, 0.293242443818599, 0.970651295036077, 0.186052078614011, 0.924251572228968, 0.199407777981833, 0.979701755801216, 0.43423207802698, 0.963663005968556, 0.863485085312277, 0.825896560912952, 0.458889417117462, 0.840306138619781, 0.314600150566548, 0.32141851214692, 0.178468223894015, 0.82039353181608, 0.721240092767403, 0.74876881018281, 0.48360933503136, 0.696837144671008, 0.122258440591395, 0.873603118816391, 0.287862420314923, 0.86360476189293, 0.509197373408824, 0.052990805124864, 0.127769862767309, 0.389307867502794, 0.271020979387686, 0.450252360897139, 0.480486689833924, 0.261814515339211, 0.7560270561371, 0.619641165714711, 0.290773614076898, 0.23756845202297, 0.245157311670482, 0.220450476743281, 0.419469280168414, 0.49378570751287, 0.27358627342619, 0.450866979779676, 0.948710192460567, 0.409109304426238, 0.602988169994205, 0.364767900202423, 0.696973968762904, 0.567818405572325, 0.412431910866871, 0.768386424519122, 0.805778163485229, 0.235286907292902, 0.318864794913679, 0.862858669366688, 0.431708253221586, 0.580874585779384, 0.29049005266279, 0.898900505853817, 0.0830053384415805, 0.854160443181172, 0.428473386447877, 0.890942008467391, 0.186899886233732, 0.585255157900974, 0.359189546434209, 0.847281507216394, 0.30053698620759, 0.354604898951948, 0.899683663388714, 0.121120858471841, 0.764647068455815, 0.418304783524945, 0.662902026437223, 0.604278746759519, 0.466860082000494, 0.176098132738844, 0.403718108776957, 0.913369797402993, 0.577616779599339, 0.965976656647399, 0.700027342885733, 0.129697000840679, 0.352329524001107, 0.0191356432624161, 0.359294122783467, 0.283226931933314, 0.95681122969836, 0.737479841336608, 0.832755861338228, 0.604845389723778, 0.0821564318612218, 0.705537574365735, 0.208165450021625, 0.779034920036793, 0.608700121287256, 0.0976104515139014, 0.00200896104797721, 0.887815569294617, 0.168209361610934, 0.0238445203285664, 0.37244299496524, 0.0774160493165255, 0.284485969459638, 0.368334453785792, 0.641177083132789, 0.0312280824873596, 0.735311876051128, 0.614558736095205, 0.29168732627295, 0.18677309830673, 0.0424211234785616, 0.556431849021465, 0.00412303721532226, 0.196868314407766, 0.851941826753318, 0.637659688247368, 0.937517965678126, 0.452387117547914, 0.2783599330578, 0.898674827767536, 0.910537716699764, 0.890723624266684, 0.267936352873221, 0.719067768193781, 0.907401184085757, 0.779023579088971, 0.688019418157637, 0.371112906839699, 0.824486881727353, 0.565698511432856, 0.00782994413748384, 0.358255746774375, 0.0892764800228179, 0.467014812864363, 0.489840853959322, 0.446391565958038, 0.426008783048019, 0.594981587026268, 0.857939029345289, 0.614006062503904, 0.740984943928197, 0.885546752484515, 0.504807833116502, 0.514673882862553, 0.374275679467246, 0.839905619388446, 0.185590433888137, 0.0908922792878002, 0.231059844372794, 0.298256011912599, 0.374500280246139, 0.0518829505890608, 0.794977670069784, 0.55105658993125, 0.517824388109148, 0.0467658033594489, 0.494348041946068, 0.59050828195177, 0.478868522681296, 0.0276048951782286, 0.972965700551867, 0.774176395731047, 0.438066631089896, 0.851065594470128, 0.018871168140322, 0.0694004436954856, 0.0698120936285704, 0.958129025995731, 0.529809533851221, 0.50012649060227, 0.171486283186823, 0.0897121226880699, 0.632384206168354, 0.353808315703645, 0.12245571706444, 0.665561725851148, 0.994101704796776, 0.169525534845889, 0.853354818886146, 0.967709759948775, 0.406968050636351, 0.136672116350383, 0.733279147418216, 0.331687558442354, 0.63407678459771, 0.964901017956436, 0.173019220819697, 0.375478162197396, 0.45518512558192, 0.730261836666614, 0.708653893088922, 0.00893628527410328, 0.261814867379144, 0.00529060885310173, 0.511484967777506, 0.336701328633353, 0.977929659886286, 0.48154209763743, 0.0846908779349178, 0.529884305084124, 0.453569172183052, 0.767980344127864, 0.55138489883393, 0.216288424329832, 0.254185401601717, 0.692903929390013, 0.17291668523103, 0.829431764548644, 0.791656734189019, 0.430190269369632, 0.356184158241376, 0.737738640978932, 0.561628793366253, 0.448880071984604, 0.812695705331862, 0.498830927303061, 0.526196063961834, 0.244819768704474, 0.865642796736211, 0.338340223999694, 0.173093998804688, 0.0118414212483913, 0.139213051181287, 0.684744303347543, 0.723886378109455, 0.138744974741712, 0.253736671758816, 0.136517015518621, 0.470556017709896, 0.268170048948377, 0.294325683265924, 0.0529782143421471, 0.763295490294695, 0.884552020812407, 0.673949430696666, 0.972141730133444, 0.470831722486764, 0.756397950695828, 0.923020360758528, 0.603907668497413, 0.745129847200587, 0.965829897206277, 0.583304376108572, 0.280324140097946, 0.0436882330104709, 0.257219310849905, 0.0402516182512045, 0.899671665858477, 0.242000626865774, 0.199396025389433, 0.68117974488996, 0.825676363660023, 0.351102139567956, 0.702926101628691, 0.272080412134528, 0.050245373044163, 0.880106825381517, 0.433906282531098, 0.0684149088338017, 0.734421076951548, 0.495432291645557, 0.468243078328669, 0.579390890197828, 0.309121191268787, 0.836842021439224, 0.305985498474911, 0.287334982305765, 0.362420401303098, 0.604403427569196, 0.47574219526723, 0.999554218258709, 0.177328075282276, 0.656848484650254, 0.23434815579094, 0.58789182943292, 0.86545429029502, 0.188110021874309, 0.368604938266799, 0.832037198357284, 0.343195907073095, 0.655400317627937, 0.0224496608134359, 0.215130288153887, 0.312316815601662, 0.754048394039273, 0.534942813916132, 0.186545167351142, 0.706202811095864, 0.15248872875236, 0.170590125024319, 0.0379471643827856, 0.4166115578264, 0.107766652246937, 0.668551810085773, 0.4397785323672, 0.316057496704161, 0.889910747297108, 0.633759409189224, 0.378327015321702, 0.433984376955777, 0.0125559912994504, 0.98845406761393, 0.824916428187862, 0.980067672906443, 0.270578752504662, 0.633621490793303, 0.909121912671253, 0.531171529786661, 0.6967347136233, 0.78918186109513, 0.536539460299537, 0.889790796907619, 0.118246800266206, 0.797848234185949, 0.403767347335815, 0.693287562346086, 0.0931464063469321, 0.000329993665218353, 0.477707965765148, 0.879032543860376, 0.106462256517261, 0.638171830913052, 0.206970871193334, 0.36695801327005, 0.160397148923948, 0.931780857034028, 0.449444924015552, 0.748948521446437, 0.908772903960198, 0.884277946781367, 0.520095706218854, 0.819554038811475, 0.835478027584031, 0.325176784303039, 0.23786487383768, 0.417092721909285, 0.772777965292335, 0.0762133868411183, 0.408282139338553, 0.791414017323405, 0.0295122510287911, 0.720779628958553, 0.0803971684072167, 0.964428510284051, 0.179858785122633, 0.439677829155698, 0.0106864813715219, 0.86058379127644, 0.410375494975597, 0.238417761167511, 0.798503981437534, 0.314065069425851, 0.655649479478598, 0.0996788076590747, 0.297879740130156, 0.0661136375274509, 0.102553575765342, 0.952031774679199, 0.687229326926172, 0.119464553892612, 0.532781326211989, 0.511889818124473, 0.885456932941452, 0.362824602751061, 0.304428504547104, 0.729093173053116, 0.119039415381849, 0.251992870587856, 0.917937957448885, 0.879847578005865, 0.821211434900761, 0.444123982451856, 0.260723354062065, 0.141258576884866, 0.998331730253994, 0.129361853003502, 0.123065117979422, 0.71702505950816, 0.177862670039758, 0.905468381941319, 0.119209727738053, 0.917779754847288, 0.789476189296693, 0.235245401505381, 0.84952542395331, 0.35019932128489, 0.151254424126819, 0.0259156611282378, 0.070136743132025, 0.298450250411406, 0.148598681204021, 0.409846232738346, 0.518791719805449, 0.493487360887229, 0.151496754726395, 0.813076050020754, 0.875920632854104, 0.350353093584999, 0.107008039718494, 0.456210287055001, 0.382524409797043, 0.729907744331285, 0.132897204719484, 0.457679926650599, 0.549914463190362, 0.570143235847354, 0.607206974877045, 0.738394221290946, 0.328665282577276, 0.188538514310494, 0.211329803336412, 0.742212679935619, 0.538669047178701, 0.271672124508768, 0.450993457343429, 0.182112880749628, 0.233515620464459, 0.75312035670504, 0.834010692313313, 0.47434498812072, 0.679709344636649, 0.731491607381031, 0.00697640981525183, 0.264653474325314, 0.557652624323964, 0.949499777518213, 0.959193347487599, 0.317467207554728, 0.879586570896208, 0.987773830536753, 0.725955343572423, 0.43981049535796, 0.194097231840715, 0.371255978709087, 0.778376647969708, 0.0714732168707997, 0.707884538918734, 0.600356182781979, 0.306455245241523, 0.724589186254889, 0.852876862278208, 0.41051207180135, 0.143559092422947, 0.415754359448329, 0.696195437107235, 0.732118483632803, 0.548226232174784, 0.527417815290391, 0.565061986912042, 0.204679999733344, 0.95377079048194, 0.958460826659575, 0.20756335137412, 0.0073892546351999, 0.538134069880471, 0.897352833300829, 0.327848595101386, 0.501973377540708, 0.476770004956052, 0.748897225130349, 0.62403830466792, 0.0556953088380396, 0.495771753834561, 0.40578549169004, 0.663596935337409, 0.719120324356481, 0.329307687934488, 0.184486476238817, 0.439573723357171, 0.737533204257488, 0.238774724770337, 0.30438967444934, 0.662722061853856, 0.509096187073737, 0.180233307415619, 0.0561298506800085, 0.479617716511711, 0.0931096829008311, 0.558580479584634, 0.40806662500836, 0.207971809664741, 0.808105335803702, 0.720465564867482, 0.327308109495789, 0.109028084436432, 0.619410161161795, 0.856074080104008, 0.769200131762773, 0.437565659871325, 0.336408275179565, 0.39530302840285, 0.737534667598084, 0.720052310498431, 0.418702772818506, 0.734046947211027, 0.11794143426232, 0.320004717214033, 0.981080531841144, 0.439297630218789, 0.832955898949876, 0.0776234930381179, 0.889140411978588, 0.912243793020025, 0.554493027739227, 0.130457294872031, 0.546092195436358, 0.8042849348858, 0.646023162640631, 0.829116159817204, 0.224523241166025, 0.175443839980289, 0.0352813631761819, 0.281089034397155, 0.543212027987465, 0.030975129455328, 0.345479313051328, 0.866702296072617, 0.0688256751745939, 0.219035766320303, 0.33986386586912, 0.90963462064974, 0.047230965225026, 0.758858551038429, 0.73546321853064, 0.709803524427116, 0.231727388221771, 0.14727149088867, 0.580826578196138, 0.111115447012708, 0.205562696559355, 0.347832702100277, 0.0731154542881995, 0.198915001470596, 0.0899386561941355, 0.415616011014208, 0.511936938157305, 0.918680357281119, 0.974651138298213, 0.625958558870479, 0.645989863201976, 0.595022828318179, 0.137308434117585, 0.199847387615591, 0.983612083829939, 0.399918642826378, 0.505635573994368, 0.350792594021186, 0.738636692520231, 0.255749279865995, 0.940899042179808, 0.158415368758142, 0.0800603958778083, 0.294100128347054, 0.791985431453213, 0.96762579632923, 0.646932796109468, 0.619340015109628, 0.864404860883951, 0.120613286737353, 0.946633086772636, 0.588623382616788, 0.882576057454571, 0.674665253376588, 0.78454849217087, 0.565647289389744, 0.0057424244005233, 0.411175164626911, 0.898863641079515, 0.576373480493203, 0.0844013609457761, 0.776029949774966, 0.893147711409256, 0.11871607019566, 0.581331137567759, 0.821350029902533, 0.804678010754287, 0.191437480039895, 0.617132338695228, 0.908029131824151, 0.410070847254246, 0.444875830318779, 0.385962206404656, 0.97545493254438, 0.53850840870291, 0.760182961123064, 0.315138783771545, 0.0863573227543384, 0.268032629974186, 0.0384360251482576, 0.550721464445814, 0.519035137956962, 0.850742721231654, 0.496660251636058, 0.425175895914435, 0.276656579691917, 0.690942633664235, 0.612771378364414, 0.593325941124931, 0.830263479379937, 0.798897352069616, 0.31402720999904, 0.675714234123006, 0.415173198329285, 0.995544753968716, 0.523789793485776, 0.49939888343215, 0.0314849088899791, 0.183598804986104, 0.723152418853715, 0.570081805577502, 0.32408056454733, 0.392200604546815, 0.345800537616014, 0.951902687782422, 0.740270990412682, 0.0554888378828764, 0.565735521027818, 0.59358568280004, 0.985323005355895, 0.489297984400764, 0.3172233463265, 0.176559814251959, 0.150110539980233, 0.774722244590521, 0.21093448321335, 0.188287206692621, 0.296543909003958, 0.832220185780898, 0.278354877838865, 0.830420669633895, 0.48320773197338, 0.843448891304433, 0.280958288582042, 0.83141070464626, 0.255116560729221, 0.874533030902967, 0.990102679235861, 0.806416846578941, 0.938178361859173, 0.210058154771104, 0.270423632813618, 0.845787957310677, 0.468330807285383, 0.990404535084963, 0.979216627543792, 0.434967653825879, 0.696455992991105, 0.240799603285268, 0.00734236789867282, 0.493319841101766, 0.633264948613942, 0.076803935226053, 0.834674335084856, 0.991437166463584, 0.157709813211113, 0.988395721418783, 0.329623523866758, 0.877827797085047, 0.0837405086494982, 0.870388504117727, 0.103236183524132, 0.489755727117881, 0.326898401835933, 0.212454021908343, 0.981658716453239, 0.283903754549101, 0.939947306877002, 0.64797172578983, 0.899692232254893, 0.548032874241471, 0.0366256434936076, 0.829830435104668, 0.804997263941914, 0.0811211101245135, 0.24018621025607, 0.761917984345928, 0.0303869431372732, 0.474563977215439, 0.0666477011982352), actual = c(0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.536983603145927, 0.424636285984889, 0.96119057177566, 0.574344806373119, 0.142189689911902, 0.880073161562905, 0.193887929897755, 0.146634917473421, 0.720008832402527, 0.58605948719196, 0.839528506854549, 0.966822555288672, 0.143847844796255, 0.484475136268884, 0.0477948621846735, 0.537566383834928, 0.869958990020677, 0.0547124224249274, 0.138990821316838, 0.732020778348669, 0.74985094787553, 0.921699115540832, 0.805067093344405, 0.990555811440572, 0.137342782923952, 0.469294481445104, 0.970516566885635, 0.247227949555963, 0.14648539875634, 0.118814607616514, 0.452120947651565, 0.39565687882714, 0.985022128093988, 0.783029852667823, 0.867168230004609, 0.58130063675344, 0.0327683165669441, 0.986836603377014, 0.927674689795822, 0.290287488605827, 0.949769687606022, 0.0733198928646743, 0.205019944580272, 0.579598870826885, 0.673934076912701, 0.538668879540637, 0.971512753050774, 0.889551181811839, 0.659384215716273, 0.915334625169635, 0.236674497136846, 0.597369941417128, 0.88391457637772, 0.256063311360776, 0.502899014856666, 0.938808091450483, 0.31188867916353, 0.404598056338727, 0.234249736648053, 0.195601047482342, 0.794161684345454, 0.0363608731422573, 0.412872542161494, 0.258723314385861, 0.982520434539765, 0.0166488715913147, 0.522693634033203, 0.739730986766517, 0.117447953904048, 0.395402782829478, 0.206912199268118, 0.680622994434088, 0.262810331303626, 0.721662111580372, 0.495332568185404, 0.209902303060517, 0.578724617604166, 0.960967929568142, 0.797557214740664, 0.440697352169082, 0.886175107909366, 0.447994093876332, 0.853760184952989, 0.564496655482799, 0.528536308789626, 0.459143881220371, 0.246486216317862, 0.163608441362157, 0.519432335393503, 0.0810850609559566, 0.206700915237889, 0.333811402553692, 0.0624626190401614, 0.478212313493714, 0.467497110599652, 0.266160008730367, 0.0603398766834289, 0.555623378837481, 0.477226920891553, 0.382787668611854, 0.671416575089097, 0.349476201459765, 0.85363219701685, 0.364629457471892, 0.565125840017572, 0.0781634389422834, 0.168682979419827, 0.329394564032555, 0.413049009395763, 0.3582848100923, 0.901911320863292, 0.357036568922922, 0.990652403794229, 0.708722573006526, 0.54740687366575, 0.630836178781465, 0.520926166558638, 0.665315556107089, 0.0205750558525324, 0.187362300464883, 0.665613644523546, 0.196139535168186, 0.315620836336166, 0.854801755864173, 0.92539797630161, 0.817834542365745, 0.19017651793547, 0.500274024670944, 0.739577139960602, 0.295892758760601, 0.373910522321239, 0.594329535029829, 0.513882225146517, 0.687060349853709, 0.861913452623412, 0.727237895829603, 0.895112411817536, 0.922243777429685, 0.0679176945704967, 0.0572794075123966, 0.122803087579086, 0.537010855972767, 0.136038357391953, 0.362021318404004, 0.770389175741002, 0.702168000396341, 0.627392818918452, 0.473089426755905, 0.589029655558988, 0.11882418137975, 0.657257097773254, 0.95174392266199, 0.392903712810948, 0.959876652806997, 0.840599536430091, 0.324635244673118, 0.110955349402502, 0.318560220999643, 0.349939600331709, 0.00721558486111462, 0.767373526468873, 0.500099466880783, 0.605357642052695, 0.5576304087881, 0.503159203799441, 0.379639795282856, 0.374906212091446, 0.654126570792869, 0.796853018691763, 0.218499803682789, 0.771995077840984, 0.543233000207692, 0.730445420136675, 0.561234356835485, 0.548714719479904, 0.534742967691272, 0.101017806213349, 0.83751143142581, 0.580858125118539, 0.207520629046485, 0.982668302720413, 0.896970284869894, 0.0330268626566976, 0.0500694704242051, 0.989459176082164, 0.283980051986873, 0.0113852340728045, 0.0955563017632812, 0.12690917798318, 0.500262193148956, 0.0665328807663172, 0.956707334611565, 0.282899938058108, 0.0505079948343337, 0.562302910257131, 0.157292959280312, 0.526954633649439, 0.00602527218870819, 0.506252276711166, 0.218159309355542, 0.998678207397461, 0.461301533039659, 0.693354904884472, 0.660372033715248, 0.854947450105101, 0.0769016123376787, 0.15966684021987, 0.997359791537747, 0.40855858894065, 0.656106592388824, 0.0278089493513107, 0.0142889423295856, 0.550070770084858, 0.487851844634861, 0.69294904358685, 0.773534846957773, 0.618724670726806, 0.995697727194056, 0.897908452665433, 0.212164732860401, 0.742873444920406, 0.699546534568071, 0.250758782029152, 0.880075752269477, 0.202447331976146, 0.733981483848765, 0.848371266620234, 0.00612997869029641, 0.458851166768, 0.320511150872335, 0.508232329506427, 0.621900020167232, 0.219087599078193, 0.308426911244169, 0.768933313433081, 0.760001447517425, 0.308382007526234, 0.0237468194682151, 0.692578639602289, 0.470606868853793, 0.955219028051943, 0.559870407218114, 0.300294577842578, 0.911861810134724, 0.99890456860885, 0.984462431864813, 0.894599922932684, 0.799782043322921, 0.602728698635474, 0.297612365568057, 0.301067038439214, 0.62900245655328, 0.765344641404226, 0.354582387953997, 0.108194461558014, 0.0689149431418628, 0.0937498332932591, 0.929663048824295, 0.241567401215434, 0.995479552540928, 0.0599806460086256, 0.272054064786062, 0.905007595196366, 0.710734396474436, 0.440256430301815, 0.504726093029603, 0.119323745602742, 0.821979727596045, 0.636138096218929, 0.506514007225633, 0.277459979522973, 0.83583102026023, 0.78224163223058, 0.492260369937867, 0.219789685448632, 0.969363447278738, 0.596012240508571, 0.201254057465121, 0.672653441084549, 0.985264570685104, 0.441930215805769, 0.325110592879355, 0.876984871458262, 0.419062501983717, 0.312065859558061, 0.460607776651159, 0.153573167743161, 0.515681371558458, 0.81047652522102, 0.540610509924591, 0.292199110379443, 0.530584555352107, 0.60541986534372, 0.266222837148234, 0.350942282937467, 0.284012925345451, 0.238333323504776, 0.61980202049017, 0.0246964001562446, 0.00842176540754735, 0.856069410452619, 0.308333697961643, 0.066303197061643, 0.640144667122513, 0.207524269120768, 0.263525383779779, 0.84468091907911, 0.492788184201345, 0.711524729849771, 0.804583217250183, 0.386643451172858, 0.869271929142997, 0.441284546162933, 0.221917250659317, 0.80252477270551, 0.685622399440035, 0.290545093361288, 0.146952911280096, 0.853615863481537, 0.331577012781054, 0.484820732148364, 0.12713651265949, 0.115314065478742, 0.621840962208807, 0.153813328593969, 0.221931287553161, 0.08495371369645, 0.020430346718058, 0.765346445841715, 0.0768565030302852, 0.788075418910012, 0.966522498987615, 0.00448553450405598, 0.622173047158867, 0.392694670008495, 0.664465631823987, 0.223155212355778, 0.447864043293521, 0.631489643128589, 0.638148710830137, 0.8744097573217, 0.156482322141528, 0.39524651132524, 0.214112844783813, 0.274102942785248, 0.683640631148592, 0.83032663166523, 0.718031229451299, 0.0477441665716469, 0.893702417379245, 0.670652362052351, 0.241602927213535, 0.118789267959073, 0.340585253899917, 0.151171231176704, 0.470589136704803, 0.31852568523027, 0.0935079506598413, 0.0239241907838732, 0.12300782604143, 0.160582593642175, 0.0790210526902229, 0.156890416983515, 0.278001909144223, 0.389921441674232, 0.814431597944349, 0.578810296021402, 0.921949053648859, 0.889953575562686, 0.135052121710032, 0.624385949689895, 0.158565387828276, 0.720808532554656, 0.38626989768818, 0.626705431146547, 0.402617514599115, 0.396246034884825, 0.0862036088947207, 0.451911768643185, 0.97766620805487, 0.149370904546231, 0.359648261219263, 0.19754530233331, 0.530417969683185, 0.0734985736198723, 0.291165244532749, 0.214738345704973, 0.537174712866545, 0.632641640258953, 0.839949177578092, 0.300472946139053, 0.427020568866283, 0.603683635592461, 0.737569865072146, 0.120228474959731, 0.251506843836978, 0.9466582974419, 0.996103539131582, 0.509831094415858, 0.365612100576982, 0.855336834676564, 0.214632661314681, 0.155130878323689, 0.776374110020697, 0.121049497742206, 0.368598269298673, 0.788277033250779, 0.932059466838837, 0.575516439508647, 0.494948605075479, 0.624672604026273, 0.546588612254709, 0.115625177742913, 0.82402933598496, 0.926847285358235, 0.934744360623881, 0.65690694283694, 0.0519499208312482, 0.727486588526517, 0.676300521241501, 0.453597357962281, 0.598033711314201, 0.685669995145872, 0.338081555673853, 0.583884730003774, 0.744955271948129, 0.0450281705707312, 0.363744664238766, 0.479536278871819, 0.96513114310801, 0.980640584370121, 0.457891910336912, 0.913071701303124, 0.73053164430894, 0.14207945461385, 0.623767109354958, 0.00338002294301987, 0.454176199156791, 0.620134200202301, 0.0360421726945788, 0.897252337541431, 0.793318739393726, 0.717487019952387, 0.950326093705371, 0.522969242418185, 0.884572095703334, 0.737182278884575, 0.588789079105482, 0.31163620040752, 0.564711423590779, 0.12096650339663, 0.360731569118798, 0.92765512666665, 0.275657169753686, 0.884270448237658, 0.277551366947591, 0.217751534190029, 0.533614507177845, 0.807229839963838, 0.351463790750131, 0.492057033581659, 0.39279416622594, 0.201468828832731, 0.484654417494312, 0.434665836161003, 0.224830580642447, 0.1661855308339, 0.291158899664879, 0.566596353892237, 0.0670777456834912, 0.037012473680079, 0.621410555671901, 0.531027761055157, 0.390234989346936, 0.867948354454711, 0.634838478872553, 0.428214080631733, 0.736725080059841, 0.284997679060325, 0.204472691053525, 0.504191077779979, 0.41448519914411, 0.702253495110199, 0.913075714372098, 0.422288128640503, 0.450454918900505, 0.814232720294967, 0.513331774156541, 0.748377045383677, 0.185637579066679, 0.842292489018291, 0.0710173540282995, 0.0820372714661062, 0.597997859120369, 0.775261335773394, 0.766708070877939, 0.199420516844839, 0.0844638908747584, 0.732347910292447, 0.744669705163687, 0.523571854457259, 0.956038998672739, 0.613518852507696, 0.465877819573507, 0.62158628879115, 0.983801950700581, 0.158085918985307, 0.210274305660278, 0.804019556147978, 0.794970076996833, 0.914411129429936, 0.262163770850748, 0.329136542975903, 0.732841729884967, 0.587900958955288, 0.487277172040194, 0.364924462977797, 0.632171622943133, 0.809692355804145, 0.853895555716008, 0.477328483248129, 0.236827459651977, 0.635017735417932, 0.735850017983466, 0.70161117776297, 0.229402894619852, 0.960029923124239, 0.752855019178241, 0.253016641596332, 0.554452212527394, 0.326152839232236, 0.850842710118741, 0.802726013818756, 0.443732316140085, 0.749861113261431, 0.99587528523989, 0.370120335137472, 0.366379259387031, 0.770036833127961, 0.392369055422023, 0.243168440880254, 0.689529894385487, 0.868433825671673, 0.0413871058262885, 0.645361218368635, 0.130244814790785, 0.304775525117293, 0.00839328905567527, 0.667979140765965, 0.0979149336926639, 0.37283771764487, 0.0543341070879251, 0.878399162553251, 0.793817335274071, 0.889219432370737, 0.332983056548983, 0.0420199257787317, 0.0595719225239009, 0.177047109929845, 0.548685214947909, 0.410361208720133, 0.557574486127123, 0.506182030541822, 0.793395843589678, 0.0929024524521083, 0.809145960491151, 0.229758635163307, 0.438583481358364, 0.993130036862567, 0.973850484471768, 0.0618137260898948, 0.0867241697851568, 0.156946174800396, 0.990427341777831, 0.662232326343656, 0.227121681906283, 0.668573746457696, 0.533096262952313, 0.639087044401094, 0.353544660145417, 0.0734220943413675, 0.227104032412171, 0.532374750124291, 0.221064006211236, 0.049403531011194, 0.467275490518659, 0.533242983743548, 0.523735178168863, 0.33569417684339, 0.373343672836199, 0.677642534021288, 0.966233658604324, 0.593310690484941, 0.831828521331772, 0.80114814825356, 0.0912898578681052, 0.398170033935457, 0.432835785206407, 0.554550146684051, 0.142883625812829, 0.0415731873363256, 0.345435156254098, 0.386080786585808, 0.565486242761835, 0.449448760366067, 0.273479450494051, 0.14402414043434, 0.522200242383406, 0.761140982387587, 0.886724826181307, 0.355840023374185, 0.924247958231717, 0.685766008682549, 0.477444390300661, 0.137832905631512, 0.556621466530487, 0.808808889705688, 0.107721124310046, 0.0880510825663805, 0.61724859662354, 0.553689909167588, 0.46229087933898, 0.283856480382383, 0.939946790691465, 0.634119072696194, 0.480256352806464, 0.0323018899653107, 0.611101577756926, 0.99782258621417, 0.791819893289357, 0.598107871832326, 0.947210453916341, 0.852511023171246, 0.253277335315943, 0.197261908324435, 0.0594653021544218, 0.141811510315165, 0.755189972231165, 0.00232365587726235, 0.736975579988211, 0.187304614344612, 0.894579229177907, 0.920207771006972, 0.303327683592215, 0.470171580789611, 0.593007614370435, 0.424114753492177, 0.323867501225322, 0.790262082824484, 0.155475318897516, 0.901288745459169, 0.0533811706118286, 0.570698458468542, 0.728697108570486, 0.873609363799915, 0.317623279057443, 0.618068549782038, 0.227514717495069, 0.809861022979021, 0.651326789753512, 0.141402394743636, 0.710193599341437, 0.415110442088917, 0.967468531802297, 0.302536379545927, 0.759813755750656, 0.322181251831353, 0.0466591429430991, 0.679675878025591, 0.32886072457768, 0.587942721555009, 0.523777524242178, 0.3997210431844, 0.84277980052866, 0.855017870897427, 0.986233701230958, 0.76991627481766, 0.638719115639105, 0.191514192149043, 0.051316759781912, 0.870922532863915, 0.573847040068358, 0.282644264632836, 0.373818733496591, 0.954381888499483, 0.783218247117475, 0.469791870564222, 0.984360625501722, 0.387888726079836, 0.573006763588637, 0.0129939401522279, 0.362369383685291, 0.603437267011032, 0.107795045478269, 0.828313816338778, 0.21390137844719, 0.0586820431053638, 0.722285912372172, 0.413126244908199, 0.443468575831503, 0.158565665129572, 0.42638747766614, 0.129745327634737, 0.65564062888734, 0.815971156582236, 0.111768667353317, 0.818250490585342, 0.671546394005418, 0.10093800467439, 0.171045662136748, 0.66348898713477, 0.609856443013996, 0.641277534188703, 0.684180947719142, 0.613109307130799, 0.96398607455194, 0.244361903285608, 0.309400161262602, 0.0627359454520047, 0.52355074766092, 0.644321415340528, 0.946868686471134, 0.813319696579129, 0.139706979738548, 0.742346044396982, 0.232797003351152, 0.142917917110026, 0.53550412831828, 0.245989226270467, 0.898179769283161, 0.6243796646595, 0.0793220368213952, 0.502741330070421, 0.653600902762264, 0.412133595440537, 0.864506316836923, 0.985865948256105, 0.0279444546904415, 0.528474026825279, 0.144969147630036, 0.732050668215379, 0.0614484909456223, 0.368768437532708, 0.207995452452451, 0.548730753129348, 0.315282112220302, 0.887426572153345, 0.10489315749146, 0.135858083842322, 0.549587287008762, 0.998351181391627, 0.326618794817477, 0.384018597193062, 0.0837368783541024, 0.235000733518973, 0.386782874120399, 0.803163758711889, 0.481939967023209, 0.655953080626205, 0.908950419165194, 0.969050378771499, 0.577822409104556, 0.968959843972698, 0.505434900056571, 0.683970597106963, 0.926915659103543, 0.289104849332944, 0.497629008954391, 0.238352267304435, 0.805082889739424, 0.675579762551934, 0.69102653930895, 0.279960374347866, 0.0493748388253152, 0.369169843383133, 0.616386713227257, 0.362292294157669, 0.728617073968053, 0.870654591126367, 0.146424261387438, 0.0773203505668789, 0.007813127245754, 0.471534700365737, 0.0265241372399032, 0.390712229534984, 0.367490090662614, 0.440110165625811, 0.855623699724674, 0.576505757402629, 0.000469095539301634, 0.693845293484628, 0.176673982758075, 0.962014203891158, 0.357624947093427, 0.88759041018784, 0.856525072362274, 0.204672566615045, 0.366691343253478, 0.381608931114897, 0.109221140388399, 0.298959608655423, 0.804793320596218, 0.757552871247754, 0.992651384789497, 0.293142993701622, 0.714069179957733, 0.694246820174158, 0.14405753207393, 0.835694960318506, 0.661734921392053, 0.430857238825411, 0.370636221021414, 0.0757752684876323, 0.209523355821148, 0.599411306669936, 0.753406531643122, 0.651849654270336, 0.654133283998817, 0.628412654157728, 0.413159466115758, 0.900844426825643, 0.498268155381083, 0.803641812177375, 0.438730876892805, 0.609765081200749, 0.682349108392373, 0.464007743168622, 0.512825039215386, 0.740407125325873, 0.132313689216971, 0.0138940506149083, 0.135649414965883, 0.216514972737059, 0.107705387519673, 0.813168913358822, 0.852296669501811, 0.0185015129391104, 0.0628860550932586, 0.298424602951854, 0.413913272088394, 0.668398446636274, 0.713772516231984, 0.352823247201741, 0.937852289061993, 0.637395633151755, 0.281959700863808, 0.912618396570906, 0.54976336681284, 0.86774941557087, 0.558517199708149, 0.0321023603901267, 0.840714231831953, 0.109216091688722, 0.898136148462072, 0.0715601556003094, 0.403576593846083, 0.423166109481826, 0.651663776719943, 0.119261709041893, 0.837018347112462, 0.0889434264972806, 0.515363476937637, 0.451666091335937, 0.721367257880047, 0.811849022749811, 0.810717500047758, 0.891683989902958, 0.151351440465078, 0.123099860269576, 0.238947679754347, 0.486740904860198, 0.840475378325209, 0.0973209517542273, 0.728726725792512, 0.169818297261372, 0.909690190339461, 0.494957621209323, 0.779522754251957, 0.531651781639084, 0.532175146741793, 0.13789490936324, 0.655365593964234, 0.111249327426776, 0.37448779749684, 0.644807748030871, 0.697119226679206, 0.458642319077626, 0.953491197666153, 0.676022330764681, 0.636431790655479, 0.0877234046347439, 0.797085488913581, 0.476709316018969, 0.467471326002851, 0.0974661547224969, 0.415093407034874, 0.094988418277353, 0.941133534768596, 0.267395156435668, 0.195483461255208, 0.421008056961, 0.398997246986255, 0.170356938848272, 0.192110051400959, 0.374115717830136, 0.866105200722814, 0.242462156107649, 0.599827780388296, 0.692431835224852, 0.964620946673676, 0.492136990651488, 0.640276819933206, 0.0393203464336693, 0.933744784677401, 0.0608495860360563, 0.205297692213207, 0.128764444729313, 0.81515442347154, 0.00128715066239238, 0.720909735187888, 0.688225823221728, 0.921346680494025, 0.971641320735216, 0.0477917375974357, 0.751433979952708, 0.649409491801634, 0.694258868927136, 0.36177101614885, 0.0672871835995466, 0.266318514710292, 0.0198140372522175, 0.569503103615716, 0.967313340865076, 0.447363898158073, 0.900699813617393, 0.451858589192852, 0.566909921588376, 0.0269061082508415, 0.771893194876611, 0.204531580442563, 0.440194311086088, 0.121392023283988, 0.513439445756376, 0.587244212860242, 0.939485579030588, 0.865193015197292, 0.631186838028952, 0.89224463631399, 0.2375244833529, 0.419465988641605, 0.101330424891785, 0.776774344732985, 0.24715857510455, 0.140312429284677, 0.98018459463492, 0.943297297228128, 0.809230499435216, 0.322761215735227, 0.822094843024388, 0.243494638940319, 0.274845346109942, 0.580261153634638, 0.370508424472064, 0.203779800096527, 0.920969963539392, 0.262621615780517, 0.687708620680496, 0.360944211483002, 0.244765653042123, 0.623586278641596, 0.782463299343362, 0.96619743318297, 0.374693701043725, 0.999118077103049, 0.0256830877624452, 0.967298248084262, 0.433352436870337, 0.154997961828485, 0.204798453487456, 0.406029894715175, 0.366821193369105, 0.517398753203452, 0.272260769736022, 0.343137532938272, 0.459213715745136, 0.631964339874685, 0.758907921379432, 0.657771917525679, 0.530467643169686, 0.681465561501682, 0.482693552738056, 0.160905401455238, 0.821943828370422, 0.0468753431923687, 0.3764942670241, 0.750917112221941, 0.0451480713672936, 0.732347742654383, 0.508374501019716, 0.271601865068078), actual = c(0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"