Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.9912369        0.1780914 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5123958        0.5123958        0.4946111 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.0726367563474923, 0.873134564841166, 0.895870692795143, 0.21623425045982, 0.82263566297479, 0.807125106453896, 0.268015163950622, 0.142124851234257, 0.320254576625302, 0.365449970588088, 0.862627064576373, 0.871744762640446, 0.970620911568403, 0.471481486689299, 0.446244347142056, 0.391941733192652, 0.471863298444077, 0.125647379551083, 0.921831390121952, 0.204169274540618, 0.585648721316829, 0.953343544853851, 0.576165583450347, 0.123625622829422, 0.352260323707014, 0.817958299769089, 0.173598242923617, 0.832898834953085, 0.728209343273193, 0.32455546874553, 0.783388237236068, 0.182947300374508, 0.959660200867802, 0.324193098582327, 0.917226374149323, 0.357950333505869, 0.0395477481652051, 0.888167649041861, 0.00647843000479043, 0.43966194964014, 0.781402430962771, 0.894479693844914, 0.801108895568177, 0.714025683468208, 0.974682269385085, 0.829326302744448, 0.0847102722618729, 0.309290835401043, 0.0301744285970926, 0.908241326920688, 0.234854834154248, 0.490409976802766, 0.665145544800907, 0.438578680390492, 0.64551779977046, 0.106642363592982, 0.614051715703681, 0.0879199367482215, 0.634476473554969, 0.689485537353903, 0.815064834430814, 0.83177265804261, 0.545737165724859, 0.564078712603077, 0.534050367539749, 0.761533513199538, 0.0566338230855763, 0.0392941068857908, 0.0878436455968767, 0.838746496010572, 0.745109414448962, 0.782373770140111, 0.00930135394446552, 0.402180666802451, 0.194857801543549, 0.715157034806907, 0.372641638154164, 0.701086369575933, 0.551585137145594, 0.389743178384379, 0.753825846128166, 0.778323423583061, 0.934599484549835, 0.881665813270956, 0.932123021688312, 0.44158825231716, 0.721659687813371, 0.837415588786826, 0.011152246966958, 0.17249219189398, 0.307263627415523, 0.885513532906771, 0.797218839172274, 0.486656252061948, 0.642114594345912, 0.290937632089481, 0.33819470833987, 0.323857218725607, 0.463689316064119, 0.145763518055901, 0.847366824978963, 0.928874950390309, 0.376489142421633, 0.489597600884736, 0.518104516202584, 0.140531215583906, 0.808154222089797, 0.451098250225186, 0.814566467190161, 0.639820502372459, 0.357162474887446, 0.59714957466349, 0.394054051255807, 0.0457190191373229, 0.20888379169628, 0.535852309782058, 0.984970792429522, 0.0544023173861206, 0.42479439615272, 0.0683055152185261, 0.363438126863912, 0.742628168547526, 0.30942180682905, 0.536847405601293, 0.623613745439798, 0.986361362040043, 0.355893024709076, 0.639197281328961, 0.026928351027891, 0.884489580988884, 0.730440798448399, 0.774736454244703, 0.259909927845001, 0.380724942777306, 0.274861474754289, 0.738221925450489, 0.161281949840486, 0.568389252759516, 0.950657784705982, 0.679484182503074, 0.338232070207596, 0.365675903856754, 0.0561704880092293, 0.69173748511821, 0.308879743563011, 0.0706345299258828, 0.388987898593768, 0.542004883522168, 0.564877019962296, 0.569195142714307, 0.421873448416591, 0.967991838930175, 0.41913735633716, 0.198356810258701, 0.509242312517017, 0.0748451256658882, 0.851995496079326, 0.853964260546491, 0.137439991347492, 0.429741964209825, 0.465860409662127, 0.00246493914164603, 0.174231366720051, 0.147025592392311, 0.219700050773099, 0.201234614709392, 0.63524052943103, 0.892922427039593, 0.736785950371996, 0.630797238554806, 0.980657495092601, 0.185646890895441, 0.239532026229426, 0.161845321534202, 0.488323442172259, 0.25572507828474, 0.758787348400801, 0.843460549134761, 0.457752370741218, 0.0127073773182929, 0.288465461693704, 0.889779324643314, 0.974541338859126, 0.557998562930152, 0.723320536781102, 0.429225945612416, 0.963849198771641, 0.266630233963951, 0.633920044871047, 0.55809282977134, 0.929829736705869, 0.110411833738908, 0.837355435825884, 0.613861584570259, 0.382690922357142, 0.683606831822544, 0.903991693630815, 0.846360811265185, 0.911937014665455, 0.931069169659168, 0.861197128891945, 0.489223659969866, 0.853442053543404, 0.0777984960004687, 0.482416148995981, 0.251301789656281, 0.31092760222964, 0.612035803264007, 0.394674505572766, 0.975190924713388, 0.595862422138453, 0.296847858699039, 0.108598182443529, 0.143726795911789, 0.0880076345056295, 0.414015714544803, 0.0059680100530386, 0.751236429903656, 0.833852445939556, 0.977432979969308, 0.805236618965864, 0.0639642588794231, 0.342234233161435, 0.973482374101877, 0.854282105807215, 0.571398302912712, 0.368203236022964, 0.554502410348505, 0.784923487575725, 0.978369368473068, 0.335067049367353, 0.979518237290904, 0.836168167879805, 0.688922793371603, 0.46438986598514, 0.559450152795762, 0.696550430962816, 0.663253091974184, 0.108266407623887, 0.711515373084694, 0.686845821561292, 0.467357988702133, 0.497236094903201, 0.288659534184262, 0.423911802703515, 0.705058161169291, 0.744380295975134, 0.869520712876692, 0.359664438758045, 0.155284883687273, 0.991348554380238, 0.259470072574914, 0.0926910096313804, 0.391202094964683, 0.224756194744259, 0.832100165309384, 0.762868979945779, 0.633362343301997, 0.250722731463611, 0.0743431670125574, 0.574609765550122, 0.126531706890091, 0.371047218795866, 0.738929518731311, 0.813711717491969, 0.704642004799098, 0.694352155784145, 0.314416043227538, 0.240084159187973, 0.993622610578313, 0.0961817461065948, 0.655854454729706, 0.528206236194819, 0.0163491703569889, 0.291195936733857, 0.237128934357315, 0.778241127263755, 0.500388768967241, 0.250644077314064, 0.114134129602462, 0.795971597312018, 0.914040663512424, 0.693340034689754, 0.840758565114811, 0.406929387710989, 0.0641966285184026, 0.195540070068091, 0.959299849811941, 0.743466686923057, 0.227836739970371, 0.0074418792501092, 0.279606458265334, 0.372919471003115, 0.660533310612664, 0.914129484212026, 0.839830187615007, 0.302090947516263, 0.383325394941494, 0.404423323227093, 0.742991752922535, 0.324607006972656, 0.103696119738743, 0.603137562284246, 0.581318006617948, 0.00130970869213343, 0.371475767111406, 0.594410908874124, 0.0369877298362553, 0.636741456575692, 0.735211083898321, 0.0946004446595907, 0.729210157413036, 0.617845287313685, 0.400737376883626, 0.823289882857352, 0.123751975595951, 0.261890611611307, 0.929272361565381, 0.409817543346435, 0.741558055626228, 0.19745321245864, 0.074939307756722, 0.495916700921953, 0.322707849089056, 0.119343067053705, 0.889624659903347, 0.454790974734351, 0.106865076581016, 0.421350804856047, 0.268558500800282, 0.542418247787282, 0.592231676448137, 0.465914961881936, 0.634117860114202, 0.535151489311829, 0.864536416949704, 0.894246922340244, 0.358531949110329, 0.170770656550303, 0.305874046171084, 0.790063043590635, 0.0717367136385292, 0.427059764508158, 0.0451666216831654, 0.786639862926677, 0.449419473996386, 0.442330041667446, 0.662846995517612, 0.0423054550774395, 0.288997246650979, 0.0460681554395705, 0.57908379402943, 0.340933110332116, 0.533773554256186, 0.375132554676384, 0.859657544875517, 0.337264490779489, 0.715583950746804, 0.0243140766397119, 0.682610924821347, 0.840590429957956, 0.668322037905455, 0.108197443885729, 0.0602165646851063, 0.73656479595229, 0.212292247451842, 0.336127077229321, 0.430427111918107, 0.715666360221803, 0.992936526192352, 0.397155019687489, 0.898169605759904, 0.607768470887095, 0.830778931733221, 0.943881037179381, 0.0290542142465711, 0.315194217022508, 0.249670706456527, 0.314698424423113, 0.241127600427717, 0.849849389167503, 0.889715550234541, 0.132996626431122, 0.151001092744991, 0.0312349509913474, 0.794642564374954, 0.208281659753993, 0.0447015671525151, 0.620214447379112, 0.946403112728149, 0.393777448218316, 0.0349066921044141, 0.319978166604415, 0.986920347902924, 0.249193246709183, 0.242815127596259, 0.173080750042573, 0.997712560463697, 0.195151468971744, 0.656044710194692, 0.539441718021408, 0.197208260651678, 0.0394008816219866, 0.96227105287835, 0.97075114864856, 0.620707078604028, 0.373114977497607, 0.559566875454038, 0.313274289248511, 0.20838891970925, 0.821502798935398, 0.226436174241826, 0.453080125851557, 0.416995529318228, 0.965950564015657, 0.740381097188219, 0.733495868043974, 0.986316413152963, 0.81619775458239, 0.590651296311989, 0.537984990514815, 0.885591493221, 0.92982359486632, 0.434008526150137, 0.920376812573522, 0.208976945606992, 0.532577211502939, 0.459940053988248, 0.105111525859684, 0.469876680290326, 0.971407288918272, 0.433507482521236, 0.557492026127875, 0.171819885028526, 0.971377826761454, 0.280006626620889, 0.40757891559042, 0.569862445350736, 0.926200377522036, 0.404970376985148, 0.187280303332955, 0.217713580699638, 0.69503111904487, 0.211681471439078, 0.350291242357343, 0.131245032418519, 0.694229145534337, 0.973361425101757, 0.447924262611195, 0.874255427159369, 0.384089480852708, 0.230547171086073, 0.438071901211515, 0.249927228083834, 0.533020365983248, 0.264241671189666, 0.226026866119355, 0.646654482232407, 0.0299074440263212, 0.967221562750638, 0.0924395525362343, 0.194196519674733, 0.0547503917478025, 0.10999419237487, 0.315766772022471, 0.748721909476444, 0.304582955315709, 0.689110714942217, 0.200844632694498, 0.960268434602767, 0.138926550745964, 0.991578920744359, 0.401183036388829, 0.582778576295823, 0.744262394262478, 0.232982240850106, 0.966831972124055, 0.974060037871823, 0.227118416922167, 0.345578358508646, 0.845093371579424, 0.251206138869748, 0.28535717073828, 0.860638353275135, 0.436181134078652, 0.734874189132825, 0.980089844670147, 0.664810516871512, 0.88934097182937, 0.686284694820642, 0.365686723031104, 0.897823467850685, 0.0495081483386457, 0.0851832826156169, 0.0454352227970958, 0.812649741070345, 0.17598470300436, 0.417128814617172, 0.243865553755313, 0.947376121534035, 0.366016466636211, 0.132224408909678, 0.955487199127674, 0.255837318021804, 0.834982547909021, 0.0958326729014516, 0.939040912780911, 0.077198983868584, 0.167993469862267, 0.962605607695878, 0.700225413078442, 0.647109067300335, 0.663772259606048, 0.629879733780399, 0.867819695267826, 0.533175689401105, 0.0023957728408277, 0.673011127160862, 0.518983720103279, 0.0176751606632024, 0.187320860801265, 0.771237690467387, 0.407105482416227, 0.744333676993847, 0.536632209317759, 0.968683283776045, 0.167192250490189, 0.851645210990682, 0.486103281145915, 0.0520671000704169, 0.521785879042, 0.833143620053306, 0.528557611396536, 0.162006812402979, 0.563097478589043, 0.0122509256470948, 0.771447605919093, 0.777823636773974, 0.78346659289673, 0.631712971720845, 0.469972792314366, 0.623095787130296, 0.0136735050473362, 0.0562874642200768, 0.294947063550353, 0.424243397545069, 0.0746968335006386, 0.512684682151303, 0.790678347926587, 0.317627431591973, 0.0842985115014017, 0.0879361489787698, 0.0308607737533748, 0.675257866503671, 0.870908464072272, 0.346031543100253, 0.621013500262052, 0.0583530173171312, 0.62255626055412, 0.260912013472989, 0.688910134835169, 0.77765459776856, 0.601618553511798, 0.965032593812793, 0.810172330820933, 0.734909673919901, 0.805326181231067, 0.0653344918973744, 0.453011708334088, 0.653791208053008, 0.456458185100928, 0.288122814847156, 0.603495477931574, 0.197484403150156, 0.474092960823327, 0.245277355890721, 0.703538797097281, 0.478788519045338, 0.158622657880187, 0.929151342250407, 0.865544275846332, 0.431907212827355, 0.342240469297394, 0.416674259351566, 0.184687485685572, 0.463693737052381, 0.411896748933941, 0.982514598174021, 0.529493847396225, 0.135004153707996, 0.836377463070676, 0.858526679454371, 0.00930559122934937, 0.346111442428082, 0.557724424637854, 0.402061385335401, 0.878071821760386, 0.115820016944781, 0.963937603170052, 0.790888154646382, 0.558565600542352, 0.822348716901615, 0.501798157580197, 0.123602144652978, 0.744070821907371, 0.317089330870658, 0.98977101710625, 0.976353822974488, 0.694710853043944, 0.374740072526038, 0.532898660050705, 0.0472080674953759, 0.212635181844234, 0.499082313850522, 0.602853524265811, 0.639750722795725, 0.836025823140517, 0.103305397555232, 0.579847377724946, 0.0974995284341276, 0.677975789178163, 0.688903836300597, 0.637213354930282, 0.784620916703716, 0.29197372472845, 0.356339783873409, 0.784490566700697, 0.676283686887473, 0.69889435172081, 0.206052157096565, 0.0914416569285095, 0.486859399825335, 0.0406875116750598, 0.830546797486022, 0.856230059405789, 0.382183716865256, 0.347523229429498, 0.482622682815418, 0.801094794180244, 0.175855096429586, 0.0167444448452443, 0.727827645372599, 0.811355267185718, 0.862902502762154, 0.917456500697881, 0.677174341399223, 0.222823034971952, 0.791469837771729, 0.035085751209408, 0.0491713115479797, 0.977595353033394, 0.543583947932348, 0.630141789093614, 0.385612301295623, 0.00334532954730093, 0.381546174641699, 0.695108572952449, 0.0338469576090574, 0.164565723156556, 0.431495150784031, 0.811541812494397, 0.787576581351459, 0.866289711557329, 0.558521854924038, 0.722319949185476, 0.433450665092096, 0.173616386018693, 0.622878070920706, 0.973940041847527, 0.228482002858073, 0.760575984138995, 0.484758777543902, 0.0283682090230286, 0.353362426161766, 0.568655653158203, 0.522700443398207, 0.817538569914177, 0.24507079902105, 0.895978821907192, 0.782075414666906, 0.638928562169895, 0.554972091689706, 0.0801385894883424, 0.747790849301964, 0.762019027024508, 0.250375545350835, 0.0621725912205875, 0.0863924331497401, 0.467083868570626, 0.922165428288281, 0.949116988573223, 0.611744099063799, 0.458357483614236, 0.258705192245543, 0.949649682734162, 0.910801096819341, 0.958561732433736, 0.220231637125835, 0.837312100455165, 0.090414366684854, 0.970696287229657, 0.281309934798628, 0.849168978398666, 0.125388155225664, 0.689507938921452, 0.625220857560635, 0.531080020358786, 0.494505437556654, 0.0213575044181198, 0.355168666690588, 0.536039516329765, 0.0476629703771323, 0.189715251326561, 0.490816853009164, 0.015399583382532, 0.0272847185842693, 0.664295741124079, 0.0774576899129897, 0.0504921653773636, 0.780281748622656, 0.258351548342034, 0.413889193674549, 0.0910667057614774, 0.813444707775488, 0.569103094050661, 0.0844200844876468, 0.455420130630955, 0.0210003857500851, 0.734398332890123, 0.190257714828476, 0.361721175024286, 0.121774253202602, 0.985306489048526, 0.457532910630107, 0.521936547243968, 0.381395157659426, 0.600116379326209, 0.732220937497914, 0.635371658718213, 0.714197117369622, 0.114528015488759, 0.552720537874848, 0.450573648093268, 0.183184408349916, 0.457077593542635, 0.728340194094926, 0.649212897988036, 0.206686582416296, 0.839983806479722, 0.289913918357342, 0.504004691261798, 0.363180484389886, 0.283711644588038, 0.563573627499864, 0.00976025080308318, 0.795934754190966, 0.222851531812921, 0.366655652411282, 0.402461181394756, 0.599983045132831, 0.568371349247172, 0.349731337744743, 0.211178446887061, 0.176394198089838, 0.590299591422081, 0.840245444560423, 0.824270050739869, 0.32734597264789, 0.861749364528805, 0.889585737837479, 0.560716173378751, 0.653628594474867, 0.715252513997257, 0.705386519664899, 0.718289804877713, 0.958664966048673, 0.590748193208128, 0.629542356356978, 0.35561048425734, 0.900839999085292, 0.615326552884653, 0.191537517821416, 0.86294007441029, 0.335838904138654, 0.67262446321547, 0.615401468705386, 0.844727971823886, 0.379072256851941, 0.0657119089737535, 0.651389466132969, 0.127768623875454, 0.634592334507033, 0.832436022581533, 0.0985907709691674, 0.257195515325293, 0.885174540337175, 0.898073698626831, 0.541443217312917, 0.606492475606501, 0.605423466535285, 0.778821606654674, 0.68852022732608, 0.0691514248028398, 0.189652527915314, 0.215131821110845, 0.172674525529146, 0.0808082348667085, 0.398533805971965, 0.0101508270017803, 0.270706429379061, 0.503521760227159, 0.482594901230186, 0.818708992563188, 0.497464469168335, 0.477705243742093, 0.49592199572362, 0.764769068919122, 0.301561490632594, 0.975035580573604, 0.0537090778816491, 0.976995291188359, 0.82648700196296, 0.827981723472476, 0.851109531475231, 0.977943802950904, 0.315315301995724, 0.213900918141007, 0.244053977308795, 0.479046194348484, 0.296988917281851, 0.25296017783694, 0.673630310920998, 0.0500526027753949, 0.0600478784181178, 0.553761383984238, 0.923912806436419, 0.878700640751049, 0.875825862865895, 0.354656559182331, 0.5922420269344, 0.236604778794572, 0.745414808625355, 0.76158477505669, 0.981288688257337, 0.366042724344879, 0.269703835947439, 0.429590197978541, 0.97323131817393, 0.865211255382746, 0.691878365818411, 0.25965232308954, 0.878062400501221, 0.201706704217941, 0.958324044942856, 0.437718579545617, 0.898406279738992, 0.936100900173187, 0.222329383948818, 0.232687247218564, 0.692956112558022, 0.473663476062939, 0.501487504923716, 0.951539270346984, 0.949580613756552, 0.675714815268293, 0.379652450094, 2.30616424232721e-05, 0.538027494912967, 0.558736698236316, 0.759416383923963, 0.292054794263095, 0.482510626781732, 0.0259608293417841, 0.335980707779527, 0.451007748255506, 0.145506694447249, 0.791744767455384, 0.787767421454191, 0.0100015765056014, 0.342756518861279, 0.444953778525814, 0.843869452597573, 0.0664717075414956, 0.642555117374286, 0.0835058046504855, 0.671285967342556, 0.709926878334954, 0.267546724760905, 0.913602931657806, 0.686161412391812, 0.0540444790385664, 0.409985646605492, 0.863838628865778, 0.199465790996328, 0.167024732800201, 0.188657913822681, 0.21335544413887, 0.504526761360466, 0.795316333416849, 0.927128451876342, 0.99195434525609, 0.750644235638902, 0.654434602707624, 0.96670446661301, 0.332185473758727, 0.163851258344948, 0.537044069496915, 0.593654438387603, 0.962341212667525, 0.506683921208605, 0.210633944021538, 0.747660835972056, 0.355260234791785, 0.600163046969101, 0.381237307796255, 0.174277595011517, 0.209552800515667, 0.00878530112095177, 0.813577022403479, 0.854920506477356, 0.503076310269535, 0.0114162969402969, 0.365798875689507, 0.122977509628981, 0.45204009138979, 0.528323601698503, 0.701892289333045, 0.318133943481371, 0.528238071361557, 0.284846373600885, 0.629880581982434, 0.465179525082931, 0.147035783156753, 0.528688689228147, 0.305923033040017, 0.050613789120689, 0.0997606343589723, 0.400885814568028, 0.740897370735183, 0.298434017924592, 0.059333567507565, 0.125524295959622, 0.580253209685907, 0.82222275226377, 0.0764918127097189, 0.269575094105676, 0.941566302673891, 0.653650001157075, 0.551791205536574, 0.886551399482414, 0.401471790159121, 0.214294365141541, 0.0354228282812983, 0.396126005332917, 0.486893810564652, 0.405004174914211, 0.389683304121718, 0.939894709270447, 0.332241567550227, 0.151874962029979, 0.154905429575592, 0.280539679341018, 0.0913450850639492, 0.737683453131467, 0.185370010556653, 0.780325534055009, 0.0698465362656862, 0.991302757058293, 0.157732218038291, 0.168967921286821, 0.944300015922636, 0.551209030672908, 0.326748556690291, 0.516480597667396, 0.276029608212411, 0.00890814093872905, 0.594610058236867, 0.240305604645982, 0.675210998393595, 0.510830747429281, 0.832559484057128, 0.969656125176698, 0.219481665641069, 0.862589166266844, 0.540620686719194, 0.904911758610979, 0.0793575856368989, 0.542412578128278, 0.57141349860467, 0.137917902320623, 0.623634974472225, 0.764087485382333, 0.759241539752111, 0.103917550994083, 0.639311436098069, 0.911104849772528, 0.962491345359012, 0.0236807917244732, 0.339573202421889, 0.438216092996299), actual = c(1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.557155599351972, 0.656316150678322, 0.212172746891156, 0.14246641844511, 0.263227140065283, 0.722372811520472, 0.339066210668534, 0.444183155195788, 0.979808498639613, 0.540367628447711, 0.37859861459583, 0.985292751109228, 0.284817700041458, 0.407917434815317, 0.372852973872796, 0.712941702920943, 0.48253733292222, 0.501679038628936, 0.130883142584935, 0.354546951828524, 0.446347765158862, 0.938584507908672, 0.290887748356909, 0.129933595424518, 0.973776066675782, 0.0796745086554438, 0.853016495704651, 0.229947800980881, 0.990507056005299, 0.417279698653147, 0.389385549584404, 0.304597544716671, 0.823870820226148, 0.383036548271775, 0.17208614712581, 0.801174324704334, 0.591545841889456, 0.194861353142187, 0.41703700996004, 0.862005979986861, 0.410013591405004, 0.768745448440313, 0.479371440131217, 0.614516705507413, 0.428519395878538, 0.733738290611655, 0.240147087490186, 0.445612138137221, 0.380476219579577, 0.538213727530092, 0.257761698449031, 0.902899119770154, 0.0942101627588272, 0.49846950545907, 0.476869803853333, 0.366610966157168, 0.310924362391233, 0.285282980883494, 0.715753721771762, 0.609425205737352, 0.924705160781741, 0.590326582081616, 0.499477130593732, 0.0602092633489519, 0.560984916053712, 0.130874339491129, 0.708038727752864, 0.96700852829963, 0.938206648686901, 0.0861002390738577, 0.193819815758616, 0.886643587145954, 0.213002238189802, 0.241132466122508, 0.416015642229468, 0.0400817384943366, 0.84599703992717, 0.105165123473853, 0.775177713716403, 0.90819089487195, 0.214647392276675, 0.0922324080020189, 0.97831854200922, 0.79525958513841, 0.728931806050241, 0.599688229383901, 0.245843886164948, 0.698689467972144, 0.366702191764489, 0.83045573416166, 0.844854401657358, 0.636060817865655, 0.594575084513053, 0.854334078961983, 0.163624683627859, 0.455240337410942, 0.425547871971503, 0.961637010099366, 0.817491049179807, 0.29240453010425, 0.994496000232175, 0.419197597075254, 0.585461095441133, 0.366234524175525, 0.604301161365584, 0.365380632691085, 0.655998785980046, 0.309050439856946, 0.126469318754971, 0.146300605963916, 0.466654043644667, 0.664716680301353, 0.0403498569503427, 0.813898690277711, 0.753169840201735, 0.505861300509423, 0.590530880261213, 0.171899895416573, 0.46149959997274, 0.309473179047927, 0.892414276255295, 0.524298102129251, 0.579616193659604, 0.32673630816862, 0.419085029978305, 0.351399055216461, 0.386234349803999, 0.592629753053188, 0.108297016704455, 0.391965156188235, 0.733941497514024, 0.862828781362623, 0.546143479412422, 0.812250360613689, 0.782728114631027, 0.573413875885308, 0.283760425634682, 0.987898823106661, 0.255351684754714, 0.700254395371303, 0.216361001832411, 0.357523652957752, 0.183935484383255, 0.573925427626818, 0.207492060726508, 0.383546418510377, 0.679061856353655, 0.0576980903279036, 0.903501397231594, 0.204119181726128, 0.472643692512065, 0.980679439380765, 0.0902426086831838, 0.687595266848803, 0.270774165866897, 0.877511069877073, 0.936627042014152, 0.234367769444361, 0.241622772766277, 0.463124635862187, 0.0329589995089918, 0.887070790166035, 0.443796739215031, 0.574749349150807, 0.403809047769755, 0.205552982166409, 0.501177355181426, 0.793383505195379, 0.654499238124117, 0.621997977141291, 0.441539525287226, 0.368415281642228, 0.852657205658033, 0.353315664455295, 0.857615779619664, 0.780984342098236, 0.771937079494819, 0.189770511584356, 0.867055974202231, 0.665021646069363, 0.953246528049931, 0.0288794890511781, 0.78302114084363, 0.402107385452837, 0.458993406035006, 0.112591804005206, 0.393520035082474, 0.864978476194665, 0.187455175910145, 0.491446897853166, 0.92308451118879, 0.979807858122513, 0.315280939685181, 0.268865966470912, 0.229232649551705, 0.664310421328992, 0.859408222837374, 0.213999543106183, 0.588768410496414, 0.660036281915382, 0.234331432729959, 0.370727867819369, 0.875359907746315, 0.61309636826627, 0.0216311041731387, 0.455711001064628, 0.900364838773385, 0.69428902445361, 0.357915931614116, 0.781978115206584, 0.72651115199551, 0.534980824450031, 0.239125315798447, 0.0609019580297172, 0.684794341214001, 0.491709594847634, 0.0274513477925211, 0.38491354114376, 0.346102330135182, 0.882659714668989, 0.112608594121411, 0.170864268671721, 0.607339980546385, 0.156176478601992, 0.191755424020812, 0.602684243116528, 0.0387373906560242, 0.933828104753047, 0.846709688426927, 0.159426279598847, 0.391985104419291, 0.328787018777803, 0.85739158000797, 0.855474313022569, 0.816407887963578, 0.41130292462185, 0.884997021174058, 0.217280140379444, 0.721969606820494, 0.596928402781487, 0.329867561347783, 0.967635949142277, 0.594564341008663, 0.0864240343216807, 0.789162474218756, 0.383759868098423, 0.753631229279563, 0.736016181763262, 0.93738099350594, 0.124669935321435, 0.978457640158013, 0.609108339762315, 0.886383900884539, 0.16101426561363, 0.907445178600028, 0.130576229188591, 0.821562466211617, 0.644394903210923, 0.529406689573079, 0.072547524003312, 0.696556894108653, 0.951450775377452, 0.907829107483849, 0.107463355874643, 0.00350357033312321, 0.391883501550183, 0.681414651218802, 0.268277009250596, 0.313860320718959, 0.515293654287234, 0.946719394763932, 0.753495085984468, 0.481842084089294, 0.467832619091496, 0.75220356322825, 0.398157671792433, 0.731896186713129, 0.00976462592370808, 0.621218716725707, 0.569681147113442, 0.805422749137506, 0.688102693296969, 0.272630981169641, 0.975715734995902, 0.0131495664827526, 0.553297774633393, 0.210819479078054, 0.476467630360276, 0.622238341718912, 0.611569358035922, 0.374098011758178, 0.478742612991482, 0.981648310786113, 0.333649959415197, 0.349057856248692, 0.0764917565975338, 0.179354967316613, 0.664220633450896, 0.804226695094258, 0.880742632551119, 0.604897003388032, 0.358419751981273, 0.725543915759772, 0.897119615226984, 0.711712740594521, 0.898243675706908, 0.933150758501142, 0.318892667768523, 0.0230920659378171, 0.668777821352705, 0.235724679194391, 0.123806225368753, 0.498686658916995, 0.809022129280493, 0.470804553013295, 0.0118301429320127, 0.807115552015603, 0.963236059527844, 0.394495459273458, 0.753729588119313, 0.123784127878025, 0.207829251419753, 0.213371740654111, 0.590572607936338, 0.549499654211104, 0.690935023827478, 0.829058049246669, 0.544620285276324, 0.102225612848997, 0.978066023439169, 0.811494644265622, 0.00445120595395565, 0.437572793569416, 0.780538405058905, 0.331287259003147, 0.144118607975543, 0.283451120601967, 0.268266777507961, 0.937519046245143, 0.648253230843693, 0.984711475670338, 0.109341508476064, 0.945719160139561, 0.728034474421293, 0.743582770461217, 0.762824813835323, 0.404293864965439, 0.991954981116578, 0.887040005298331, 0.987489434657618, 0.672841456020251, 0.284745474578813, 0.281345038907602, 0.446040346985683, 0.544284857111052, 0.770583831239492, 0.0424311822280288, 0.335033030016348, 0.198595150606707, 0.0614532481413335, 0.372569576604292, 0.151014004601166, 0.732524836901575, 0.228338243672624, 0.334482964593917, 0.971055213129148, 0.809684011153877, 0.036720234900713, 0.766589562408626, 0.220047563314438, 0.343314326833934, 0.277435532072559, 0.868627252057195, 0.927673067431897, 0.0650999476201832, 0.758351807016879, 0.365897961892188, 0.920714583713561, 0.709287277422845, 0.540724027436227, 0.42830414394848, 0.847699123201892, 0.434402539161965, 0.196814910741523, 0.555559857282788, 0.113320556469262, 0.207521012518555, 0.330471994588152, 0.0255555503536016, 0.41512621124275, 0.611811317736283, 0.316889323526993, 0.573582398705184, 0.871527437586337, 0.519110268680379, 0.47033274336718, 0.612314363243058, 0.217998658074066, 0.375576047459617, 0.668765521841124, 0.0222222502343357, 0.564936135429889, 0.385619297390804, 0.961988314287737, 0.43010495021008, 0.797932392917573, 0.622762342449278, 0.787934730527923, 0.019396907184273, 0.237415144219995, 0.165146113839, 0.980893897125497, 0.832805211190134, 0.0939101695548743, 0.972915684105828, 0.530589614761993, 0.406754816882312, 0.491990663344041, 0.268497367156669, 0.505580793134868, 0.980533312773332, 0.821593630127609, 0.586395754944533, 0.849848480662331, 0.781406419118866, 0.898903493070975, 0.162282897392288, 0.147565218154341, 0.266085886396468, 0.830057938350365, 0.110594252822921, 0.983265230432153, 0.282222117530182, 0.533118252875283, 0.614850326208398, 0.943057559197769, 0.787936124252155, 0.227907182415947, 0.482344912365079, 0.566554939374328, 0.209368481766433, 0.489065657602623, 0.852127307793126, 0.670865847263485, 0.492339141434059, 0.459808820625767, 0.965229786001146, 0.250580610707402, 0.255584754981101, 0.815169570036232, 0.475610758876428, 0.202494814060628, 0.300559796858579, 0.337080620694906, 0.141600392060354, 0.233882044907659, 0.725260129431263, 0.363295952323824, 0.24459857842885, 0.675667453790084, 0.0887802471406758, 0.171083993744105, 0.603288816753775, 0.784345486667007, 0.151510549010709, 0.740865231491625, 0.878584723221138, 0.441308615263551, 0.156947356183082, 0.734033852582797, 0.97795866522938, 0.955624111928046, 0.379978992510587, 0.899346108315513, 0.76114409789443, 0.299131656764075, 0.163633044110611, 0.850605066400021, 0.0904768251348287, 0.980146076297387, 0.847781242802739, 0.755420965142548, 0.276467412011698, 0.536902002990246, 0.595739250537008, 0.181937202578411, 0.123616053257138, 0.284804149297997, 0.805522383889183, 0.730241562938318, 0.551994310924783, 0.53953969432041, 0.433520990889519, 0.373011927586049, 0.539630660554394, 0.730790433939546, 0.551481247646734, 0.44736160710454, 0.229348050197586, 0.844792325049639, 0.379562173504382, 0.874162396183237, 0.864922249224037, 0.932168380590156, 0.483213822124526, 0.314481182722375, 0.597473403671756, 0.54922030470334, 0.250008758157492, 0.855701939668506, 0.8165741879493, 0.481784459901974, 0.151685532880947, 0.475505717564374, 0.471053178887814, 0.187263060128316, 0.454332936089486, 0.660610069986433, 0.288631741190329, 0.835388529812917, 0.719270798377693, 0.587506802054122, 0.708468503784388, 0.353064569411799, 0.704113283893093, 0.958823110209778, 0.851134097203612, 0.639116655802354, 0.559845736948773, 0.264784578699619, 0.391613574698567, 0.282308638095856, 0.924184655304998, 0.273965623462573, 0.921020905952901, 0.534222634974867, 0.351928643183783, 0.819179067388177, 0.0632376996800303, 0.871895695105195, 0.878123103408143, 0.70788049697876, 0.775601925794035, 0.785590605344623, 0.957630969351158, 0.256743463920429, 0.908310649218038, 0.344326025573537, 0.0264057579915971, 0.391960623906925, 0.987681585829705, 0.647987256292254, 0.67238748841919, 0.165713999187574, 0.206058645620942, 0.716235829517245, 0.968141619116068, 0.358260035747662, 0.716252144891769, 0.579785593552515, 0.111520462902263, 0.725795347709209, 0.867622000863776, 0.852925386279821, 0.359106081537902, 0.677972104866058, 0.832761278841645, 0.0222630242351443, 0.00986330583691597, 0.225417016772553, 0.14269849890843, 0.786583290435374, 0.214535011444241, 0.104842533124611, 0.18195571214892, 0.785723818931729, 0.86669712793082, 0.890033667441458, 0.561623693443835, 0.59133655577898, 0.243900686036795, 0.996957871131599, 0.450571494409814, 0.855181777384132, 0.689306175103411, 0.995616149622947, 0.184934229822829, 0.473187871044502, 0.574380044592544, 0.124046340817586, 0.195234747603536, 0.0146527246106416, 0.176276524318382, 0.504632604308426, 0.563377936836332, 0.147677781293169, 0.676775201922283, 0.146677643060684, 0.563616233412176, 0.432004872709513, 0.534571913536638, 0.296877160435542, 0.523241353454068, 0.727763386443257, 0.296481319703162, 0.815936514409259, 0.806605977704749, 0.707913727732375, 0.272666375851259, 0.552359505789354, 0.828200551681221, 0.335758029483259, 0.354198656743392, 0.953139546327293, 0.378781880717725, 0.388584502972662, 0.0539014858659357, 0.631509252823889, 0.941702567273751, 0.773190177045763, 0.281517552677542, 0.587942434009165, 0.97488831798546, 0.685190694639459, 0.779948585433885, 0.273848562501371, 0.993489100132138, 0.91006592172198, 0.853534031659365, 0.800634127343073, 0.215057182591408, 0.796875525731593, 0.0328629116993397, 0.7744807950221, 0.00485118478536606, 0.397031969157979, 0.815669302362949, 0.851168541470543, 0.0802922856528312, 0.68324142252095, 0.698593230219558, 0.233794880798087, 0.700337009737268, 0.884491984033957, 0.00622759596444666, 0.0426348620094359, 0.879749427782372, 0.390487614553422, 0.903372063301504, 0.64756389032118, 0.328291148412973, 0.141986057627946, 0.289298480143771, 0.72891913075, 0.644008507253602, 0.743097791913897, 0.613617521943524, 0.46262257383205, 0.223248494323343, 0.327344155637547, 0.458114752313122, 0.812115400098264, 0.485616640886292, 0.687612918205559, 0.411226071650162, 0.0862092291936278, 0.723192032659426, 0.724836317356676, 0.451185619225726, 0.098236266290769, 0.914725754642859, 0.992812493583187, 0.964579446706921, 0.781917243730277, 0.341566217830405, 0.561433192808181, 0.195611307630315, 0.417325404006988, 0.492743320064619, 0.993807967519388, 0.685680801747367, 0.0941356173716486, 0.737787706544623, 0.248586732894182, 0.523019891930744, 0.67405988345854, 0.084865513490513, 0.036012485390529, 0.286011940566823, 0.370860002236441, 0.206888417480513, 0.0455001848749816, 0.229258190141991, 0.634962808806449, 0.820544565329328, 0.969401712762192, 0.46780353737995, 0.0758276276756078, 0.106350806541741, 0.302082342095673, 0.678851290605962, 0.546626419294626, 0.200430764583871, 0.42720557958819, 0.944759433623403, 0.0388598279096186, 0.0249379493761808, 0.525935061043128, 0.857893064850941, 0.994242788292468, 0.193863259861246, 0.764794022077695, 0.503420139895752, 0.0719397470820695, 0.472439153585583, 0.145605466095731, 0.590415579499677, 0.490678823553026, 0.947817876236513, 0.197268651332706, 0.794977513840422, 0.289839863311499, 0.845443889731541, 0.895964718889445, 0.905714926309884, 0.51435919594951, 0.52407001843676, 0.442791901296005, 0.620166214881465, 0.971840787678957, 0.829136557877064, 0.437703224597499, 0.763393001630902, 0.0485164078418165, 0.226817168528214, 0.363718963926658, 0.86051166174002, 0.226703663356602, 0.737655729986727, 0.90234321448952, 0.0591837675310671, 0.685590554727241, 0.712535985046998, 0.551493350649253, 0.113661904120818, 0.460819516563788, 0.782730809180066, 0.287944110343233, 0.542590720113367, 0.707356728147715, 0.559449962107465, 0.505687187425792, 0.889233460649848, 0.639105960493907, 0.771738324780017, 0.675053242128342, 0.277372437296435, 0.900885084178299, 0.0773003101348877, 0.940433751093224, 0.886015722760931, 0.331138396868482, 0.867762583540753, 0.8937636425253, 0.137730980757624, 0.714288171380758, 0.841366056120023, 0.278722493676469, 0.548890586243942, 0.676765304524451, 0.43180967005901, 0.811602907953784, 0.662058483809233, 0.634307487402111, 0.278637545881793, 0.203192777466029, 0.29430269729346, 0.0843917969614267, 0.761383719043806, 0.0880931545980275, 0.658093373524025, 0.134828551439568, 0.414058731403202, 0.409991425694898, 0.697362903272733, 0.306100169196725, 0.644682192010805, 0.955074219498783, 0.59929473977536, 0.121364884544164, 0.914611739339307, 0.304808455519378, 0.438181728357449, 0.573834908893332, 0.71218270319514, 0.330707748420537, 0.672070996602997, 0.465956203406677, 0.573990455595776, 0.654615644365549, 0.937733072787523, 0.747573162661865, 0.663238652050495, 0.5511372492183, 0.162085705669597, 0.833043145481497, 0.956362121040002, 0.857956139370799, 0.302646595286205, 0.256848268443719, 0.13463049964048, 0.439724798314273, 0.495016797445714, 0.198763458989561, 0.220771606080234, 0.877955720061436, 0.133303990587592, 0.833326734602451, 0.582760179881006, 0.808402558322996, 0.896343170665205, 0.246294716838747, 0.693172339815646, 0.569776575779542, 0.87955155945383, 0.39431738993153, 0.769645136548206, 0.170983453048393, 0.56204622425139, 0.633318557869643, 0.630872099427506, 0.0297994152642787, 0.726506706560031, 0.365322378696874, 0.196573029505089, 0.234711125260219, 0.0325453514233232, 0.970013378653675, 0.868122098967433, 0.506123537197709, 0.935352984350175, 0.549917235504836, 0.957126964814961, 0.477536656660959, 0.423133146483451, 0.954027012223378, 0.872220412129536, 0.226412731455639, 0.0941167653072625, 0.370961177628487, 0.484327103244141, 0.0073717541527003, 0.673457989702001, 0.318608934758231, 0.129192880820483, 0.269263340625912, 0.440777866635472, 0.718195417430252, 0.918633174151182, 0.339979111216962, 0.825783029431477, 0.506276597734541, 0.259570590220392, 0.319843423552811, 0.885273736901581, 0.553217543056235, 0.617467569187284, 0.869167386554182, 0.365017459494993, 0.318942327750847, 0.46639112662524, 0.711945310933515, 0.69775925995782, 0.26917769247666, 0.670237093232572, 0.345740587450564, 0.17799594765529, 0.930989814689383, 0.00338502298109233, 0.164514072937891, 0.617717130342498, 0.816259972285479, 0.17605891963467, 0.903072285698727, 0.16023569717072, 0.0834946888498962, 0.294482811819762, 0.107729349751025, 0.0822086702100933, 0.0710237591993064, 0.194112833356485, 0.0828651348128915, 0.905047608539462, 0.984426383161917, 0.487642799504101, 0.681292672408745, 0.709951768396422, 0.407228159951046, 0.431241087848321, 0.374306597281247, 0.544096037279814, 0.20191408181563, 0.901825369568542, 0.714820379856974, 0.314318678341806, 0.965587381040677, 0.122176499571651, 0.166526812128723, 0.840569318505004, 0.30809454456903, 0.99977720505558, 0.284075738163665, 0.496808314230293, 0.0687411369290203, 0.255716407904401, 0.906616787426174, 0.355872331885621, 0.539593412075192, 0.321573033230379, 0.00887031271122396, 0.893531036330387, 0.70767823792994, 0.388974122004583, 0.0277223596349359, 0.873191993916407, 0.0811043684370816, 0.39001082838513, 0.3330252263695, 0.36650558700785, 0.292191500775516, 0.796686636749655, 0.202574947848916, 0.795799820683897, 0.463677571387962, 0.647548250155523, 0.519377857912332, 0.45566395833157, 0.534162685275078, 0.168987684883177, 0.388576411176473, 0.0486475094221532, 0.276368658291176, 0.271108254790306, 0.581257492071018, 0.628366409800947, 0.375258052488789, 0.145590722793713, 0.832322537200525, 0.0589350149966776, 0.433910050429404, 0.675017402041703, 0.22538527729921, 0.584527532337233, 0.310122874798253, 0.842785654589534, 0.0963446167297661, 0.270829332992435, 0.265702907461673, 0.827270473819226, 0.640953634399921, 0.700624707853422, 0.466680705081671, 0.0684172478504479, 0.333125672535971, 0.568044824292883, 0.72129161702469, 0.349879786837846, 0.770928824087605, 0.251737389713526, 0.54385486850515, 0.548913268372416, 0.0296963716391474, 0.52200739714317, 0.567354916129261, 0.786542642163113, 0.0701410186011344, 0.947723367717117, 0.455518148606643, 0.364000121597201, 0.4220743435435, 0.702023090096191, 0.380757997743785, 0.756662537576631, 0.261717988410965, 0.727777482941747, 0.487471089931205, 0.801918762736022, 0.954464958282188, 0.725880410522223, 0.449265385512263, 0.443040805635974, 0.456017048098147, 0.200372954597697, 0.143993830773979, 0.142048758920282, 0.97784768557176, 0.905614273156971, 0.733912940137088, 0.625774292973801, 0.421477047726512), actual = c(0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"