Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4656213        0.5132960 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5132960        0.5132960        0.5368615 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.24411380616948, 0.78983616316691, 0.381220097653568, 0.460184746887535, 0.319906920194626, 0.542637804755941, 0.828203809913248, 0.965792404254898, 0.843270881799981, 0.439416099805385, 0.217397382482886, 0.17431748425588, 0.489162454847246, 0.775280494475737, 0.194053725339472, 0.601548980921507, 0.582491670269519, 0.461040415801108, 0.213642666349187, 0.61888848990202, 0.666226177476346, 0.888990951934829, 0.111285227583721, 0.324757020454854, 0.722823115065694, 0.257260151673108, 0.466424664249644, 0.773665295680985, 0.709237347589806, 0.772830588277429, 0.265118875773624, 0.401667898753658, 0.201637057820335, 0.352001450257376, 0.59975521126762, 0.194853110006079, 0.0449442488607019, 0.670188292162493, 0.908122074790299, 0.569774456555024, 0.268531017238274, 0.831450594821945, 0.610010357806459, 0.676767055643722, 0.212363244965672, 0.776859999401495, 0.646488671423867, 0.584815974580124, 0.631398237776011, 0.676745371427387, 0.865726791089401, 0.655861657811329, 0.123395089991391, 0.418523390078917, 0.193986717844382, 0.168325115926564, 0.981355921598151, 0.661868029506877, 0.125567227369174, 0.228301526512951, 0.908964615082368, 0.330618517939001, 0.0494677633978426, 0.152319520944729, 0.0540335667319596, 0.177373975282535, 0.818917818134651, 0.554615671746433, 0.395734558347613, 0.022628650534898, 0.802381680347025, 0.692521851509809, 0.284552090335637, 0.20171711477451, 0.619995492976159, 0.185468654381111, 0.206967699807137, 0.185887259664014, 0.897143662907183, 0.859784863423556, 0.763758303364739, 0.995525916106999, 0.244646740145981, 0.296190189197659, 0.2047489953693, 0.98142035282217, 0.458731911843643, 0.902006619377062, 0.728314645355567, 0.900193600682542, 0.959596048109233, 0.786610280862078, 0.96585787832737, 0.415016835089773, 0.89524071989581, 0.789400403620675, 0.168747278861701, 0.100645238999277, 0.312336018774658, 0.732074555940926, 0.633185037178919, 0.998163402313367, 0.883182765217498, 0.336874329019338, 0.342549061402678, 0.394873130600899, 0.397846482461318, 0.34445947711356, 0.832512221997604, 0.913714321097359, 0.63118413859047, 0.62828200054355, 0.2786263092421, 0.180704017868266, 0.434149075765163, 0.345519945723936, 0.26814204454422, 0.689181992085651, 0.107211314607412, 0.904203760437667, 0.595061333617195, 0.366532037733123, 0.371855332050472, 0.933058726601303, 0.46484948694706, 0.336593151092529, 0.894916464108974, 0.574280551169068, 0.812794291879982, 0.0775888033676893, 0.653967376099899, 0.271070997463539, 0.265731196850538, 0.842350610066205, 0.671015406958759, 0.0905263822060078, 0.628625224810094, 0.155997722409666, 0.971442111767828, 0.129781160969287, 0.393424652051181, 0.637073677266017, 0.633819679031149, 0.961944945156574, 0.424525212263688, 0.604550340445712, 0.579498833278194, 0.0112541501875967, 0.410942748654634, 0.271634043660015, 0.710395236499608, 0.44878469617106, 0.896106455707923, 0.290328925009817, 0.91540553397499, 0.249769680202007, 0.456530422670767, 0.293282587081194, 0.380715903127566, 0.445956540293992, 0.0137381271924824, 0.614922929322347, 0.696310745785013, 0.119128769962117, 0.94999827677384, 0.0600411491468549, 0.189515637699515, 0.403262164676562, 0.726097207982093, 0.00407012179493904, 0.234593087341636, 0.221880519995466, 0.327845189487562, 0.383324483409524, 0.886096671456471, 0.681594279129058, 0.404671083204448, 0.554667509160936, 0.664696072228253, 0.226251358864829, 0.196712103905156, 0.134791107615456, 0.228998693870381, 0.615581644466147, 0.975358873838559, 0.303675978211686, 0.899132392602041, 0.122600410832092, 0.741074796533212, 0.985085014952347, 0.120665517169982, 0.945845191832632, 0.0751786828041077, 0.0999631078448147, 0.138871827395633, 0.768186019267887, 0.970543057657778, 0.290195292327553, 0.423469081521034, 0.815189274260774, 0.615984061965719, 0.442096081795171, 0.984041236806661, 0.305009308969602, 0.44308235286735, 0.708944902988151, 0.93903001700528, 0.524042586795986, 0.292871316429228, 0.420264732791111, 0.609388337703422, 0.253418943844736, 0.638579195132479, 0.393020404968411, 0.354341592406854, 0.802012379048392, 0.393572549568489, 0.863841727608815, 0.519038407597691, 0.979885756503791, 0.237330436473712, 0.0637692855671048, 0.17752993805334, 0.150620175758377, 0.0635590497404337, 0.344110710313544, 0.11557479086332, 0.469952376326546, 0.0298565379343927, 0.5109407317359, 0.875472396379337, 0.364867832511663, 0.152818237664178, 0.85102861863561, 0.938416148768738, 0.0739535647444427, 0.749667320167646, 0.862873156787828, 0.129911999450997, 0.9890894619748, 0.692588709061965, 0.836588225094602, 0.451699582627043, 0.500728411832824, 0.979170487029478, 0.756027026800439, 0.867690135957673, 0.743942660279572, 0.681949687888846, 0.345120804617181, 0.841735183494166, 0.549070478184149, 0.738493319600821, 0.81975718983449, 0.61533412639983, 0.288499492220581, 0.0247062081471086, 0.721915505593643, 0.521706379950047, 0.281912897480652, 0.932619672035798, 0.728102352237329, 0.741388149093837, 0.847303247544914, 0.275798528920859, 0.671343547524884, 0.705462814774364, 0.267080793855712, 0.492837077472359, 0.460682123200968, 0.958278226433322, 0.518159357598051, 0.208273966796696, 0.00637102103792131, 0.0207759167533368, 0.311554144369438, 0.977559974649921, 0.593532059341669, 0.391003300901502, 0.988892605062574, 0.312220054678619, 0.621787551790476, 0.56314924126491, 0.924188352422789, 0.186835699947551, 0.293742215493694, 0.790900746360421, 0.734108557924628, 0.419499921845272, 0.98978716484271, 0.551149227889255, 0.426356290467083, 0.651125114411116, 0.0438168197870255, 0.000186302233487368, 0.399401161354035, 0.413875891827047, 0.737503980286419, 0.372597764711827, 0.422735666157678, 0.783973049372435, 0.0258652821648866, 0.382659364957362, 0.864601544803008, 0.298794191796333, 0.328505862504244, 0.307939705438912, 0.575150167802349, 0.907270246883854, 0.47334952862002, 0.730091927805915, 0.919241137104109, 0.415932389209047, 0.494924710132182, 0.9121012177784, 0.880024736979976, 0.524818374542519, 0.0421752787660807, 0.726393981603906, 0.82796102296561, 0.491734836949036, 0.445677876006812, 0.22477491479367, 0.551515602972358, 0.526081481948495, 0.454690166981891, 0.722999092191458, 0.0920827896334231, 0.615141479298472, 0.696418239269406, 0.0296568623743951, 0.0562764280475676, 0.817500266013667, 0.500844316091388, 0.278374190675095, 0.237933493452147, 0.0102070120628923, 0.739993065129966, 0.747867934172973, 0.888010574970394, 0.522564721293747, 0.251306030433625, 0.949961274862289, 0.303997192531824, 0.738466600654647, 0.259034062735736, 0.413494739448652, 0.211913846898824, 0.833863666513935, 0.277016251115128, 0.233083665836602, 0.557186457095668, 0.381850883830339, 0.385460056364536, 0.0745843925978988, 0.117383471690118, 0.45965593168512, 0.222617644583806, 0.060758612351492, 0.0860282317735255, 0.304487261455506, 0.32895864569582, 0.350157517939806, 0.153188294963911, 0.036662328755483, 0.411478147376329, 0.101570857688785, 0.871835988713428, 0.600420069415122, 0.117705524200574, 0.875555164646357, 0.521571235964075, 0.617135788779706, 0.505636893678457, 0.806600144365802, 0.760633926140144, 0.414167458191514, 0.203635145677254, 0.883524465607479, 0.804948589997366, 0.763985494850203, 0.620290222810581, 0.282227346440777, 0.0935535926837474, 0.131056371377781, 0.57344080414623, 0.0101026103366166, 0.640414142981172, 0.638355851173401, 0.907156977104023, 0.0389960042666644, 0.801443437114358, 0.785277048591524, 0.827889762818813, 0.451145340222865, 0.386829077964649, 0.990009600063786, 0.174896888434887, 0.786323827691376, 0.831583270803094, 0.941022427519783, 0.419994535157457, 0.832859690999612, 0.0559248488862067, 0.682658019009978, 0.308512230170891, 0.61132339364849, 0.399523612577468, 0.5662582796067, 0.966460490366444, 0.204494999023154, 0.862300292123109, 0.908145563211292, 0.177496539661661, 0.728459802689031, 0.820279374485835, 0.397106903139502, 0.86547385668382, 0.0755681565497071, 0.654457797762007, 0.594513166463003, 0.812906000996009, 0.663809160003439, 0.0705255649518222, 0.7660556754563, 0.7242264917586, 0.485995890339836, 0.223408990073949, 0.274696056498215, 0.44390057399869, 0.93231994798407, 0.535631080158055, 0.297920993762091, 0.34682573704049, 0.392234622268006, 0.4868552419357, 0.282197969267145, 0.210749837104231, 0.636041664984077, 0.930666476720944, 0.0386619376949966, 0.555082258302718, 0.742563318926841, 0.96936098835431, 0.187006842112169, 0.763332271249965, 0.392769686179236, 0.114215060370043, 0.282136181602255, 0.604868989670649, 0.308992014732212, 0.893781282007694, 0.825019874144346, 0.308587459847331, 0.667777238879353, 0.702906666323543, 0.794930666452274, 0.68515053903684, 0.041155518963933, 0.703070723451674, 0.371943681500852, 0.408493889030069, 0.558953762985766, 0.557538408786058, 0.672796092228964, 0.760725893778726, 0.917688633548096, 0.314505332149565, 0.782029221300036, 0.26292901369743, 0.709905706113204, 0.666506687179208, 0.175302535993978, 0.378315227571875, 0.534125132951885, 0.922704524826258, 0.582695368910208, 0.936677337158471, 0.759550391696393, 0.822309528011829, 0.271213138941675, 0.908239838667214, 0.450781255261973, 0.865973892388865, 0.384808220900595, 0.861634937580675, 0.524326675804332, 0.474433616967872, 0.947605994995683, 0.86432681675069, 0.375555329956114, 0.188579784007743, 0.663022208493203, 0.0909386116545647, 0.40277748205699, 0.500583372544497, 0.288782141637057, 0.976813575252891, 0.904101757798344, 0.685129446676001, 0.961904065217823, 0.494229058967903, 0.246954258298501, 0.867790134157985, 0.507383544230834, 0.251617709640414, 0.153359357034788, 0.344370993087068, 0.306823353283107, 0.700865902937949, 0.466831874102354, 0.545476217754185, 0.658115642843768, 0.942208922002465, 0.728358543012291, 0.269455471308902, 0.269035208504647, 0.158369965851307, 0.184772794134915, 0.267257711617276, 0.48041101754643, 0.591493062442169, 0.0205264575779438, 0.628202526597306, 0.0303038305137306, 0.396938242483884, 0.147703752154484, 0.357810797169805, 0.275536187225953, 0.695369569351897, 0.33367472467944, 0.502417950890958, 0.154421873157844, 0.751611658371985, 0.167338530533016, 0.287569884210825, 0.872931628488004, 0.249913221923634, 0.356778590939939, 0.591815977822989, 0.615125046344474, 0.555660422192886, 0.736566768959165, 0.217169642215595, 0.876929688034579, 0.776023266837001, 0.140419968636706, 0.835543476743624, 0.895904988516122, 0.971056364942342, 0.843409723369405, 0.0127371321432292, 0.417402962455526, 0.0894197190646082, 0.252773517742753, 0.85410484415479, 0.111089663347229, 0.329332187538967, 0.556425827788189, 0.939766712253913, 0.59485637024045, 0.300202981336042, 0.0616161399520934, 0.132382588461041, 0.33618059521541, 0.259855962358415, 0.0171719526406378, 0.883539137430489, 0.267030844464898, 0.0733816849533468, 0.206016403855756, 0.471024082042277, 0.738714310573414, 0.0690423112828285, 0.089335337979719, 0.649966332130134, 0.511199919041246, 0.58265136112459, 0.566455473424867, 0.677771795773879, 0.0859728204086423, 0.736239864490926, 0.919520756229758, 0.274667948018759, 0.883857110980898, 0.992374969646335, 0.839406341547146, 0.869584286818281, 0.0282685866113752, 0.884268678724766, 0.718366001732647, 0.653492699144408, 0.552918837638572, 0.149887296836823, 0.268011160893366, 0.881305195158347, 0.398275317391381, 0.484317839378491, 0.546923598041758, 0.0348151265643537, 0.465645617572591, 0.00907350634224713, 0.864954016869888, 0.81022757710889, 0.90146914916113, 0.536147761391476, 0.758639194769785, 0.718438589246944, 0.937612642301247, 0.931014734553173, 0.449527794029564, 0.151667919009924, 0.388661408564076, 0.985120974248275, 0.711843230761588, 0.846824694192037, 0.208521334920079, 0.513516220031306, 0.0177430706098676, 0.382949320832267, 0.663711600704119, 0.68343322025612, 0.857103104470298, 0.466643312713131, 0.903291772352532, 0.720990184228867, 0.730948691256344, 0.168488370720297, 0.490726560354233, 0.599178053671494, 0.70822444697842, 0.417468049796298, 0.225820763036609, 0.791865809820592, 0.175043763127178, 0.420990962302312, 0.890851831296459, 0.304511664668098, 0.105541140306741, 0.14360890025273, 0.263025275198743, 0.144986372208223, 0.928155754227191, 0.69023320893757, 0.245594122679904, 0.330892850179225, 0.338898135116324, 0.569835936417803, 0.372192617272958, 0.837976020760834, 0.590553134679794, 0.722260136855766, 0.430408280808479, 0.167580381967127, 0.25335058895871, 0.421504555270076, 0.408187449676916, 0.947118907235563, 0.00666581676341593, 0.902241791831329, 0.441639174474403, 0.752900621620938, 0.555354000767693, 0.961419757222757, 0.222506674937904, 0.494252559263259, 0.916385730262846, 0.912910265149549, 0.590326945763081, 0.649448370328173, 0.204486188245937, 0.48793382034637, 0.179595160298049, 0.166112887905911, 0.268462400883436, 0.673539877869189, 0.953021734021604, 0.765824926784262, 0.353007258148864, 0.268867348786443, 0.132634576177225, 0.998732095118612, 0.96688068844378, 0.883268494158983, 0.0408037703018636, 0.896303982939571, 0.581078558461741, 0.781793229980394, 0.0379008885938674, 0.239218412898481, 0.793804299784824, 0.688465394312516, 0.137730705784634, 0.480782351456583, 0.306149282027036, 0.618931490927935, 0.0273010449018329, 0.95422056876123, 0.460713357897475, 0.0841777271125466, 0.798921224894002, 0.547686064150184, 0.79715931089595, 0.659095616545528, 0.0251378775574267, 0.736383037408814, 0.640326967230067, 0.444738657446578, 0.281964744906873, 0.137856003828347, 0.36910020490177, 0.142115033464506, 0.782330539310351, 0.645429391181096, 0.452730225399137, 0.0157303407322615, 0.870476356474683, 0.71487782872282, 0.103247869294137, 0.934566588606685, 0.267120525008067, 0.381586860399693, 0.188663637731224, 0.112558186752722, 0.364405393134803, 0.146895231679082, 0.524582850281149, 0.985950329108164, 0.739914989564568, 0.317526634316891, 0.212843230459839, 0.381836374523118, 0.77802127879113, 0.0320864303503186, 0.947004921967164, 0.530886915512383, 0.203431908274069, 0.474091755691916, 0.942545607686043, 0.682125743245706, 0.637956460239366, 0.610256811138242, 0.453886873554438, 0.850365958176553, 0.12169607821852, 0.513956835493445, 0.482076251413673, 0.165382236940786, 0.338671739445999, 0.974353095516562, 0.398527146084234, 0.627165273297578, 0.303317371755838, 0.559236342553049, 0.459894311381504, 0.42625420447439, 0.0165430651977658, 0.236736227991059, 0.576990324771032, 0.950537363765761, 0.794511997140944, 0.27194985607639, 0.340541113633662, 0.884527507703751, 0.25172725552693, 0.533162005944178, 0.474637951469049, 0.648666387889534, 0.587501553818583, 0.839452496729791, 0.623655128059909, 0.714949254412204, 0.272605321137235, 0.89747915812768, 0.836602771654725, 0.106712901033461, 0.855040428927168, 0.651971845887601, 0.0879789518658072, 0.0976601850707084, 0.251634886488318, 0.213374567218125, 0.0987065327353776, 0.644211013335735, 0.558111389866099, 0.735675284639001, 0.0529762101359665, 0.469391646562144, 0.358023829059675, 0.546449718764052, 0.175821851706132, 0.328897287370637, 0.591691114706919, 0.718066621571779, 0.939735904103145, 0.615441793110222, 0.0311529480386525, 0.657770542427897, 0.580810813931748, 0.712926629697904, 0.273499509086832, 0.0828081364743412, 0.134922710712999, 0.779903013259172, 0.0646810415200889, 0.00251425267197192, 0.815266293706372, 0.726500651799142, 0.826324152294546, 0.334867525147274, 0.630922445794567, 0.884487830335274, 0.932606851682067, 0.202262722887099, 0.237654804484919, 0.577870098175481, 0.493880111025646, 0.669575338251889, 0.405423535965383, 0.237073929514736, 0.341182545060292, 0.661802901420742, 0.876738473540172, 0.449413812486455, 0.073033750988543, 0.538571366341785, 0.519728665007278, 0.69977631047368, 0.533325215335935, 0.028457130305469, 0.442718715872616, 0.0828853182028979, 0.757587031228468, 0.694252560613677, 0.94929946330376, 0.141543578356504, 0.354035673430189, 0.340468868846074, 0.133616884704679, 0.150955697055906, 0.910636266926304, 0.909877089550719, 0.117226485162973, 0.0815389144700021, 0.464327119523659, 0.0714726180303842, 0.856028181966394, 0.43518124287948, 0.0144836013205349, 0.206676776753739, 0.0970416080672294, 0.146903662709519, 0.838031281484291, 0.289820174220949, 0.565008682664484, 0.2731258708518, 0.238115788670257, 0.304932586848736, 0.0136809835676104, 0.629682731814682, 0.368345617316663, 0.202847732463852, 0.905541547341272, 0.774027368985116, 0.220416588475928, 0.416221061255783, 0.187536858720705, 0.356431900523603, 0.968036679783836, 0.778218544321135, 0.566418370930478, 0.90605772449635, 0.695259596919641, 0.100091736996546, 0.398915408179164, 0.738388518802822, 0.980544141959399, 0.0409882445819676, 0.0422009387984872, 0.928507934557274, 0.488680693553761, 0.805685356259346, 0.89963424904272, 0.0558440836612135, 0.211075579747558, 0.712644655257463, 0.0354100498370826, 0.261172192869708, 0.830308452015743, 0.304482350125909, 0.975332773523405, 0.962724917568266, 0.768124936148524, 0.0387294171378016, 0.418614550726488, 0.27607922302559, 0.831415487686172, 0.810618783114478, 0.15702666528523, 0.585347928106785, 0.526373960776255, 0.883393449243158, 0.93283834354952, 0.890094617148861, 0.103302228264511, 0.681565194157884, 0.064762209309265, 0.367190575692803, 0.883169076172635, 0.819921443937346, 0.0238475799560547, 0.774579995777458, 0.994713593274355, 0.733158668735996, 0.932561839465052, 0.52532060816884, 0.164667441509664, 0.187754536513239, 0.311321497429162, 0.218779131071642, 0.322338812518865, 0.160944670904428, 0.861445674905553, 0.567160664591938, 0.667761255055666, 0.368541477248073, 0.585810881573707, 0.149266572203487, 0.862970587564632, 0.345221284311265, 0.874942285241559, 0.109570678323507, 0.973837864818051, 0.131109453970566, 0.245140785118565, 0.812116374960169, 0.0864063119515777, 0.0198472277261317, 0.943441003793851, 0.600843549240381, 0.732033515581861, 0.0460314399097115, 0.378789910115302, 0.245430733542889, 0.24974942044355, 0.692786877043545, 0.792717093368992, 0.511401479365304, 0.555600931169465, 0.691159628564492, 0.263325615553185, 0.337986446451396, 0.187552985502407, 0.706590760499239, 0.569350640056655, 0.621535236481577, 0.613483478082344, 0.00940033816732466, 0.940913647646084, 0.66950123431161, 0.993548436556011, 0.380965909454972, 0.142896224977449, 0.0498581563588232, 0.272084157913923, 0.406182604376227, 0.956906171282753, 0.878497171215713, 0.0534672739449888, 0.40562889399007, 0.125961931655183, 0.146208926104009, 0.677459435304627, 0.716971390647814, 0.286871660966426, 0.477041609352455, 0.0363143431022763, 0.870449939509854, 0.310323006240651, 0.475733688566834, 0.659658872988075, 0.885373048949987, 0.0322313657961786, 0.562515907455236, 0.953743360703811, 0.856530496384948, 0.327973105479032, 0.349835372529924, 0.116982425097376, 0.789855825016275, 0.343790688551962, 0.860858360305429, 0.548671075375751, 0.600905400002375, 0.628627879079431, 0.110127354972064, 0.445459065958858, 0.620273154927418, 0.346390720456839, 0.395999611588195), actual = c(1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.45322616561316, 0.282571202609688, 0.527002391172573, 0.575412222882733, 0.883569836616516, 0.641814791830257, 0.517864595865831, 0.964255119673908, 0.747530102962628, 0.107672814745456, 0.837580473627895, 0.324523266404867, 0.405167629942298, 0.621446752687916, 0.768655795603991, 0.406215969007462, 0.241785230115056, 0.826075290329754, 0.696122148539871, 0.390859531937167, 0.566655008820817, 0.492646878352389, 0.328617980470881, 0.173536569578573, 0.434426996624097, 0.807176570408046, 0.479271057061851, 0.26667165919207, 0.983698358293623, 0.570804613875225, 0.318055022973567, 0.99014930985868, 0.208883205195889, 0.778987505706027, 0.166597973322496, 0.78193976264447, 0.350071786902845, 0.148543184623122, 0.370879269903526, 0.359173343749717, 0.269873441196978, 0.412728327559307, 0.520071730716154, 0.861560991033912, 0.254070726456121, 0.146975949872285, 0.435287913773209, 0.330916749779135, 0.616645212750882, 0.0872867796570063, 0.548017531866208, 0.782334832241759, 0.856898460304365, 0.199096075026318, 0.550433022901416, 0.0810861580539495, 0.871991992928088, 0.318168242461979, 0.521610740805045, 0.0652898151893169, 0.613659120397642, 0.660033030901104, 0.671870056539774, 0.961098625790328, 0.500050590606406, 0.100830705370754, 0.560978691093624, 0.909315251745284, 0.157354783965275, 0.547380832955241, 0.68808416556567, 0.0350539702922106, 0.626133745769039, 0.816394119523466, 0.543559779645875, 0.492645042482764, 0.456853003008291, 0.813255497021601, 0.304412637604401, 0.989500920288265, 0.285430274670944, 0.614497652743012, 0.534404654754326, 0.664982064859942, 0.833907746942714, 0.00274928566068411, 0.473740976303816, 0.981817518360913, 0.200293167261407, 0.657036372926086, 0.92150841630064, 0.339144726283848, 0.473142954288051, 0.828070221003145, 0.712591964751482, 0.843711158027872, 0.0192949289921671, 0.659861275926232, 0.700775074074045, 0.000212129205465317, 0.24417391512543, 0.794231317471713, 0.302926959702745, 0.348585228668526, 0.762883818941191, 0.299783144379035, 0.212620886741206, 0.679748435271904, 0.104520653374493, 0.488957928027958, 0.711430927971378, 0.332132920622826, 0.63594137574546, 0.228981551481411, 0.164379463298246, 0.753140954999253, 0.939876505173743, 0.322817547246814, 0.794489201158285, 0.344667295226827, 0.312404147582129, 0.846001736586913, 0.556903792312369, 0.116634818026796, 0.810074466047809, 0.810006311861798, 0.626849925843999, 0.342058184323832, 0.997165967244655, 0.606745799770579, 0.25974853313528, 0.477321247570217, 0.344114292878658, 0.179885935503989, 0.565050244564191, 0.880048116203398, 0.348238844890147, 0.306614991277456, 0.274813432013616, 0.666502114385366, 0.264843266457319, 0.962755431421101, 0.263106777565554, 0.788526760181412, 0.189895707648247, 0.789833299349993, 0.845396965742111, 0.128968405537307, 0.139020760310814, 0.868703278014436, 0.570167645812035, 0.559981165453792, 0.267806881340221, 0.605357585009187, 0.48747539636679, 0.221514583565295, 0.482782317791134, 0.546163647435606, 0.422495059901848, 0.558927342295647, 0.767174122389406, 0.891291379695758, 0.344707426615059, 0.259550315560773, 0.777805815450847, 0.206032428191975, 0.199867997318506, 0.436369885224849, 0.758557243971154, 0.8286055221688, 0.294332813471556, 0.586842569056898, 0.753924531163648, 0.931478319689631, 0.510633117053658, 0.204086427111179, 0.0104693670291454, 0.720028802752495, 0.936963440617546, 0.318713109474629, 0.280981381423771, 0.260926716495305, 0.270227899774909, 0.41173311881721, 0.59922056295909, 0.455147231929004, 0.0287401103414595, 0.254898726940155, 0.440625436604023, 0.967258979333565, 0.670530511299148, 0.928832965204492, 0.555883803172037, 0.455001352820545, 0.780491471523419, 0.138357985997573, 0.66873314185068, 0.678461673902348, 0.89457567082718, 0.764324350981042, 0.430126830236986, 0.62730000866577, 0.0334052075631917, 0.570505482377484, 0.103497984120622, 0.894762257579714, 0.931162142660469, 0.431308108847588, 0.476090488955379, 0.588447566144168, 0.979452266125008, 0.691157190827653, 0.688449040986598, 0.998063625302166, 0.810601839330047, 0.785123678622767, 0.584414908662438, 0.643759086728096, 0.725724113872275, 0.697961168829352, 0.0933709933888167, 0.599119069753215, 0.527671828167513, 0.975384735036641, 0.717194866621867, 0.256533936364576, 0.893714834935963, 0.563898072112352, 0.139613872626796, 0.91928336606361, 0.892492900369689, 0.217000511940569, 0.192298741778359, 0.420982697978616, 0.239206191850826, 0.860997628653422, 0.972567205317318, 0.228350908495486, 0.834576436085626, 0.146178728668019, 0.512571354163811, 0.647913477616385, 0.64184975461103, 0.478045310359448, 0.710216369945556, 0.165545499883592, 0.256313008954749, 0.820351839065552, 0.900862827897072, 0.34611471160315, 0.757628982653841, 0.903737841406837, 0.664049779763445, 0.250234863488004, 0.575115203624591, 0.273290836252272, 0.523395460331813, 0.242996698012576, 0.727794740116224, 0.546141897328198, 0.731341038132086, 0.830556615022942, 0.161704264348373, 0.993980970000848, 0.796327300136909, 0.561283559771255, 0.117755969054997, 0.868800479918718, 0.264719610800967, 0.689316330244765, 0.572834232356399, 0.659589151153341, 0.781734297517687, 0.696124677546322, 0.301940991077572, 0.199518976500258, 0.0765044963918626, 0.698001845972613, 0.441958694020286, 0.537891961168498, 0.989459135802463, 0.152509708655998, 0.520165860187262, 0.681685631163418, 0.0661560103762895, 0.490894619608298, 0.756077070254833, 0.838705578818917, 0.939837706973776, 0.380576948169619, 0.347518066642806, 0.532950073713437, 0.756607432616875, 0.0369529607705772, 0.717469442170113, 0.311028078896925, 0.0981309383641928, 0.957447075052187, 0.243290651589632, 0.634481348097324, 0.767990215448663, 0.0291709194425493, 0.836656155763194, 0.525010670768097, 0.91390418400988, 0.363133495207876, 0.0499390065670013, 0.198157265782356, 0.675403531175107, 0.943949707318097, 0.0936622347217053, 0.000519958790391684, 0.972545743454248, 0.0664937775582075, 0.384738382417709, 0.845349109731615, 0.495918770087883, 0.557488363003358, 0.295787708833814, 0.276539721293375, 0.764306961325929, 0.647872689878568, 0.221380883129314, 0.852099367417395, 0.654978148173541, 0.147249285597354, 0.27169594168663, 0.42588284984231, 0.146915599470958, 0.105067901778966, 0.669658641098067, 0.862663423409685, 0.372155428864062, 0.406347132753581, 0.0495516117662191, 0.954800987383351, 0.973005718551576, 0.0658800420351326, 0.300478106131777, 0.468791082734242, 0.792477646376938, 0.911200901260599, 0.208502420922741, 0.353994990233332, 0.259317929856479, 0.654585443902761, 0.0263937613926828, 0.118590165860951, 0.00744139333255589, 0.712113223038614, 0.501840058714151, 0.295912025030702, 0.219620516290888, 0.697210347512737, 0.444141427287832, 0.177596362074837, 0.286618282087147, 0.828332655131817, 0.575303437421098, 0.15858755656518, 0.0200093041639775, 0.139812978217378, 0.642485471209511, 0.854001940693706, 0.177354076877236, 0.0147810908965766, 0.794497586321086, 0.645268832333386, 0.809647620888427, 0.291392297716811, 0.99763695243746, 0.652127014007419, 0.059945173561573, 0.640065427171066, 0.45219737640582, 0.644991978770122, 0.175229504005983, 0.429523721104488, 0.18440380319953, 0.946875524241477, 0.386100010946393, 0.636475151404738, 0.531167406821623, 0.923197013325989, 0.767344104358926, 0.43240743689239, 0.936712831724435, 0.332365881185979, 0.888985407073051, 0.350087230093777, 0.140462121926248, 0.236352973151952, 0.925872689578682, 0.109874799847603, 0.941517744446173, 0.887801079778001, 0.432340998668224, 0.766231905203313, 0.552805611863732, 0.27919464209117, 0.722604751586914, 0.124102188041434, 0.00583538739010692, 0.0526043365243822, 0.712418221170083, 0.385129244066775, 0.499795620795339, 0.165899861371145, 0.390966176288202, 0.541764906840399, 0.363694467581809, 0.0771397133357823, 0.924038112163544, 0.654023027978837, 0.26545517751947, 0.0275084997992963, 0.76974522951059, 0.293419786961749, 0.974283224903047, 0.688846999779344, 0.612748224986717, 0.110249735647812, 0.19641361781396, 0.224242237629369, 0.184087996603921, 0.566091838292778, 0.941420465474948, 0.15322969853878, 0.912211493356153, 0.367691897554323, 0.122621040791273, 0.850104590179399, 0.746664069592953, 0.528767292154953, 0.825898063369095, 0.353313082596287, 0.968627761118114, 0.366816565627232, 0.647366726770997, 0.246271502459422, 0.257103240350261, 0.618942734319717, 0.178946947213262, 0.216539984801784, 0.619608593173325, 0.613529501948506, 0.970314632868394, 0.106688435189426, 0.662397532723844, 0.184808655874804, 0.493579261703417, 0.521544851362705, 0.325914795743302, 0.526389762759209, 0.0790780861862004, 0.709212759044021, 0.667498493101448, 0.812351120403036, 0.95635626418516, 0.499091145582497, 0.672518234001473, 0.667303472524509, 0.998026004293934, 0.653869575122371, 0.745026524877176, 0.70231219404377, 0.723861088044941, 0.0126731449272484, 0.739424458704889, 0.46051329234615, 0.25028808391653, 0.70679308171384, 0.353158494224772, 0.458814915269613, 0.899332012981176, 0.0239422805607319, 0.273199162678793, 0.595508225029334, 0.551623721374199, 0.308266833890229, 0.937540003098547, 0.984513976378366, 0.795827217632905, 0.889274677494541, 0.167122918646783, 0.145744930952787, 0.0475386229809374, 0.374827944673598, 0.50620104582049, 0.802572574932128, 0.0838892364408821, 0.990474745165557, 0.756578463595361, 0.542562208604068, 0.644009406445548, 0.113733593840152, 0.665028089424595, 0.115497257327661, 0.117365332785994, 0.706194819882512, 0.770399425644428, 0.784148108912632, 0.00251008244231343, 0.794154414674267, 0.778513999888673, 0.683717584703118, 0.15219870232977, 0.415165370795876, 0.412391076330096, 0.954931237269193, 0.944891635561362, 0.608321119798347, 0.583424698095769, 0.376543913735077, 0.444564560428262, 0.0927067643497139, 0.745595505926758, 0.904544643359259, 0.53462095418945, 0.794936394086108, 0.538712162990123, 0.267969205277041, 0.682112107984722, 0.0227907332591712, 0.054055601824075, 0.723144265357405, 0.889029585989192, 0.594554835930467, 0.420972717693076, 0.854106139391661, 0.744275197852403, 0.0742758188862354, 0.902352258330211, 0.608977993251756, 0.555016551166773, 0.0182000212371349, 0.668959234841168, 0.203141850186512, 0.775730814086273, 0.449490077793598, 0.626672830199823, 0.0692868405021727, 0.11534733325243, 0.339266758877784, 0.678346314933151, 0.396349042421207, 0.736609292216599, 0.272043507080525, 0.75170493288897, 0.252893569180742, 0.970795789035037, 0.902875633910298, 0.0666151116602123, 0.219594544265419, 0.839965528808534, 0.0177224131766707, 0.375073920469731, 0.341727284481749, 0.706004599574953, 0.52995141223073, 0.552737336372957, 0.777838720940053, 0.874742004321888, 0.39231422636658, 0.98320784047246, 0.472726283827797, 0.183305981336161, 0.309282535687089, 0.549487514188513, 0.95430077239871, 0.142777372384444, 0.209982147440314, 0.146098421188071, 0.275820204056799, 0.585615801624954, 0.915458158357069, 0.252300468273461, 0.479011822491884, 0.833189888857305, 0.245391435688362, 0.27991644362919, 0.31619037152268, 0.349517485592514, 0.746191733749583, 0.0965689080767334, 0.760641446802765, 0.0718732352834195, 0.739478763891384, 0.821222387719899, 0.0641375568229705, 0.599520933348686, 0.371546958107501, 0.597307912074029, 0.862662443891168, 0.369037605589256, 0.0531385214999318, 0.735384488711134, 0.549480492481962, 0.160676528001204, 0.882123531308025, 0.39978225203231, 0.297136131906882, 0.342976083513349, 0.0365704500582069, 0.647356989793479, 0.466660287696868, 0.897510477108881, 0.265449440805241, 0.326860796427354, 0.391361593967304, 0.381610276876017, 0.801386640639976, 0.956901378929615, 0.99856193549931, 0.921045231632888, 0.145331499166787, 0.496747683733702, 0.887460191501305, 0.893272830639035, 0.318920709192753, 0.119024969870225, 0.203660355182365, 0.850999127840623, 0.425208815606311, 0.441368631785735, 0.459963194327429, 0.903855334268883, 0.624988885130733, 0.22669987147674, 0.516731906449422, 0.536980106495321, 0.49875820800662, 0.09236308792606, 0.649904548190534, 0.872263526543975, 0.78934360621497, 0.782550184987485, 0.760864473879337, 0.48571040481329, 0.847029754659161, 0.963631081860512, 0.305327516281977, 0.879805753473192, 0.194479082943872, 0.356223594164476, 0.756134227849543, 0.898431123234332, 0.901098217349499, 0.548648236552253, 0.367352780187503, 0.910686816787347, 0.137266338802874, 0.100622438825667, 0.986765003763139, 0.411717657931149, 0.189844604814425, 0.117007382214069, 0.161407394101843, 0.282962294993922, 0.791072922991589, 0.370148388668895, 0.962751188082621, 0.0456883613951504, 0.515608686488122, 0.907363025704399, 0.848842418054119, 0.209720768500119, 0.434301150031388, 0.994123290525749, 0.870403606677428, 0.370687114074826, 0.472098045051098, 0.213937914231792, 0.937507180031389, 0.16271107760258, 0.385007135104388, 0.573472316842526, 0.0226188634987921, 0.778350008185953, 0.118044964270666, 0.526894348673522, 0.39740194263868, 0.0289602063130587, 0.379783286945894, 0.940475832670927, 0.216594459488988, 0.155454210937023, 0.931924708420411, 0.819665375864133, 0.546798444585875, 0.361217048950493, 0.972710883244872, 0.351049385033548, 0.167677532415837, 0.000322715844959021, 0.760492704808712, 0.508846279699355, 0.370013852138072, 0.929440666455775, 0.796423115534708, 0.541179422521964, 0.809964155079797, 0.97955773002468, 0.142627419205382, 0.924976020120084, 0.593542505521327, 0.121904992731288, 0.276465207571164, 0.34493018919602, 0.241011892212555, 0.0912095175590366, 0.184267955599353, 0.87008940265514, 0.182587649673223, 0.836638562846929, 0.61161708412692, 0.371438676258549, 0.432132291840389, 0.384051299886778, 0.0895761158317327, 0.67162271332927, 0.776442299596965, 0.852653234498575, 0.109240652760491, 0.259607201907784, 0.0438256994821131, 0.521499000955373, 0.905361744109541, 0.650879838038236, 0.722744334954768, 0.189700423041359, 0.934068623697385, 0.951895249309018, 0.149532455019653, 0.64390563685447, 0.950856174575165, 0.892884116619825, 0.898476452101022, 0.972260601818562, 0.312361092306674, 0.39503979915753, 0.247207541950047, 0.796371110482141, 0.770931059960276, 0.742517799371853, 0.126550604589283, 0.391356500331312, 0.26269346056506, 0.904438806232065, 0.250313029857352, 0.607246015453711, 0.63330396800302, 0.67543987906538, 0.738776570186019, 0.55877880519256, 0.841912583680823, 0.629302117507905, 0.938892923761159, 0.531702768057585, 0.258996911579743, 0.203426860040054, 0.310896537266672, 0.660163562977687, 0.764750082977116, 0.430532817495987, 0.658152500865981, 0.152158349752426, 0.317281949799508, 0.313504463993013, 0.226245119934902, 0.421584706753492, 0.145292912144214, 0.378096869215369, 0.634698888519779, 0.3865129603073, 0.144222424598411, 0.1816132562235, 0.514208048116416, 0.312494077254087, 0.570480700582266, 0.800775880925357, 0.16330086113885, 0.98878219188191, 0.722304875729606, 0.476252550492063, 0.69522351375781, 0.0613149707205594, 0.450847590574995, 0.321055136620998, 0.811008884571493, 0.465922525152564, 0.0225362232886255, 0.373597039608285, 0.988904150435701, 0.236965450225398, 0.453225412871689, 0.376813995651901, 0.175145901041105, 0.540064749773592, 0.825825612992048, 0.297620826633647, 0.416458763182163, 0.937212869990617, 0.527451771777123, 0.136030503548682, 0.424089387757704, 0.822141270851716, 0.529477830743417, 0.880867604864761, 0.590961320325732, 0.650889643467963, 0.602158086607233, 0.769211997278035, 0.426571366842836, 0.19357887422666, 0.547348810825497, 0.77055345941335, 0.20673438324593, 0.630498784128577, 0.183897890616208, 0.640288640744984, 0.742316186893731, 0.517528901807964, 0.24695090437308, 0.156950281234458, 0.677712896373123, 0.545803935965523, 0.512688154587522, 0.123585313558578, 0.438440865837038, 0.504905429435894, 0.902043510228395, 0.149017588235438, 0.354409916093573, 0.119936712551862, 0.00927816168405116, 0.199679348152131, 0.96746390638873, 0.178861345397308, 0.53514954727143, 0.86362495017238, 0.145325426245108, 0.207018166547641, 0.256786288926378, 0.922370774904266, 0.67841987661086, 0.293234632117674, 0.174892182461917, 0.738119923509657, 0.0755537149962038, 0.992052424000576, 0.042627627030015, 0.249517974210903, 0.944752773502842, 0.386788677889854, 0.777898272732273, 0.195370314642787, 0.739917467348278, 0.595149939879775, 0.760009403340518, 0.612598796607926, 0.613189732888713, 0.46899081976153, 0.709415109828115, 0.10475784400478, 0.719796736259013, 0.76408502808772, 0.752656430704519, 0.25028122891672, 0.819978622253984, 0.729734800057486, 0.227455680957064, 0.845535922097042, 0.833798414794728, 0.737343945074826, 0.0632883673533797, 0.174093632725999, 0.769682868616655, 0.552197804441676, 0.80017817299813, 0.16964187566191, 0.627801918424666, 0.478994817240164, 0.217443112516776, 0.80751544306986, 0.963285755831748, 0.948809187160805, 0.392726680031046, 0.0945010220166296, 0.791469448944554, 0.154696863610297, 0.450781228020787, 0.779511562548578, 0.492986160563305, 0.756291245808825, 0.233554994221777, 0.843449119944125, 0.13420817698352, 0.840597661677748, 0.280388539889827, 0.718197882408276, 0.904854786582291, 0.764529401902109, 0.361469793366268, 0.520730759715661, 0.263056792318821, 0.808567327912897, 0.823576406342909, 0.764276161789894, 0.752885656198487, 0.970223644981161, 0.803003857843578, 0.056389435660094, 0.704924241639674, 0.522778597660363, 0.00304799457080662, 0.136814812896773, 0.787928303005174, 0.0482071412261575, 0.898171635810286, 0.92141805565916, 0.687483449466527, 0.608843319350854, 0.748705716338009, 0.135280360002071, 0.360386429587379, 0.70680292416364, 0.682189367478713, 0.640302623622119, 0.568264953792095, 0.150888842297718, 0.663118443451822, 0.636155932443216, 0.2335332876537, 0.426180975744501, 0.171969787683338, 0.476777382194996, 0.786656228359789, 0.143625511555001, 0.0375544070266187, 0.232671168399975, 0.166544978041202, 0.652084346860647, 0.452947770943865, 0.33546643727459, 0.761624401668087, 0.544017278589308, 0.226657417835668, 0.430741964373738, 0.555851312121376, 0.597413982730359, 0.240567210596055, 0.678394271526486, 0.574282836634666, 0.892432881286368, 0.346076119458303, 0.171806775499135, 0.546901912428439, 0.195655879331753, 0.44742301478982, 0.952142773428932, 0.508765373611823, 0.667365961009637, 0.532589184120297, 0.839777365094051, 0.52775976806879, 0.615565028740093, 0.592638586880639, 0.323599918046966, 0.116531853098422, 0.677795647643507, 0.917523989221081, 0.514494109433144, 0.890336883254349, 0.961809271248057, 0.635525305755436, 0.857322848634794, 0.588028938742355, 0.499189147958532, 0.879051629221067, 0.726370682008564, 0.523735977010801, 0.524224067572504, 0.0666388340760022, 0.931061871116981, 0.709511762950569, 0.972775945905596, 0.67197408573702, 0.756837442051619, 0.71049582445994, 0.458874648204073, 0.29163809819147, 0.899452892364934, 0.096709294943139, 0.490757948718965, 0.108286982635036, 0.9858254394494, 0.948884415905923, 0.679702289868146, 0.550214070826769), actual = c(1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"