Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.1495826        0.1495826 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4924620        0.4924620        0.5009523 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.626450969837606, 0.23908102279529, 0.694742678198963, 0.111102084862068, 0.313125413609669, 0.579573845956475, 0.746567479567602, 0.174885174958035, 0.22875470854342, 0.710845011752099, 0.606871989555657, 0.0084066572599113, 0.213535874150693, 0.134021261241287, 0.618912419769913, 0.111515870783478, 0.277855687774718, 0.451948815025389, 0.0541737677995116, 0.0480107297189534, 0.262068434851244, 0.88210046524182, 0.134970673592761, 0.572472359519452, 0.0613543854560703, 0.474917208543047, 0.317192133516073, 0.0536772985942662, 0.263473640894517, 0.693496423307806, 0.66532900556922, 0.274594715796411, 0.574196066008881, 0.390065111918375, 0.875817964086309, 0.810674031032249, 0.672908175503835, 0.391188828041777, 0.327607497340068, 0.120971233816817, 0.316017867298797, 0.290517470333725, 0.159264382906258, 0.795337767340243, 0.639530266402289, 0.0725902991835028, 0.923674157587811, 0.810551668982953, 0.6483834318351, 0.413710755994543, 0.159719144226983, 0.481082393787801, 0.61748271738179, 0.149176560109481, 0.0327291218563914, 0.752180142328143, 0.670460926601663, 0.255131115671247, 0.692120217019692, 0.34024519380182, 0.319989491254091, 0.600027382839471, 0.528943755663931, 0.291871648048982, 0.998709230916575, 0.152892509708181, 0.404379759449512, 0.683699744055048, 0.135901289293543, 0.929571535205469, 0.605356827611104, 0.10357503592968, 0.649208474904299, 0.242262354353443, 0.132430125027895, 0.826710628112778, 0.0891896984539926, 0.72401978704147, 0.653858308214694, 0.0696052762214094, 0.458375509595498, 0.684541291324422, 0.341720791533589, 0.46403263323009, 0.539604943012819, 0.172435460146517, 0.499257306801155, 0.458894371753559, 0.193251427728683, 0.175521274562925, 0.902760895667598, 0.579446082701907, 0.984075953252614, 0.184755210531875, 0.239411420421675, 0.912389178294688, 0.0795631862711161, 0.564296069554985, 0.724298716522753, 0.897818831494078, 0.533338436624035, 0.270983716240153, 0.233424180187285, 0.443775420077145, 0.882830949500203, 0.0398352311458439, 0.571252665715292, 0.490647824248299, 0.252582126762718, 0.562106990721077, 0.795522536151111, 0.620420970022678, 0.0969691099599004, 0.351803445722908, 0.168451448669657, 0.299067354062572, 0.147991536650807, 0.992054056376219, 0.257876940537244, 0.394987646723166, 0.0994643452577293, 0.70358436065726, 0.405214170459658, 0.916099650086835, 0.392852117773145, 0.0357115606311709, 0.458577472250909, 0.373596767662093, 0.422371614957228, 0.718860561260954, 0.400523646501824, 0.592521980172023, 0.119641043711454, 0.060937391128391, 0.757785355206579, 0.33954931399785, 0.248357009841129, 0.494161095935851, 0.0840747943148017, 0.966342643136159, 0.881266127573326, 0.851391373202205, 0.0587741686031222, 0.439395012101158, 0.528945383382961, 0.431962421163917, 0.738369185477495, 0.00494133797474205, 0.281239261152223, 0.476102785672992, 0.994219151092693, 0.658994319848716, 0.765339480014518, 0.791388869052753, 0.42310750647448, 0.081768017495051, 0.835205956362188, 0.437869869405404, 0.796075390418991, 0.910917557775974, 0.942053192760795, 0.885623578447849, 0.448232044698671, 0.298553060740232, 0.13750920095481, 0.250410911394283, 0.170084838522598, 0.258391157723963, 0.439058408373967, 0.708413552492857, 0.263816689839587, 0.543540889862925, 0.458156684646383, 0.1604636576958, 0.454661744879559, 0.443905622931197, 0.638563637156039, 0.0285041269380599, 0.46697058621794, 0.756132624810562, 0.152603450231254, 0.605405140900984, 0.748685710364953, 0.277625101618469, 0.747464251937345, 0.281774306669831, 0.346834411611781, 0.650285890558735, 0.0153982569463551, 0.0369964297860861, 0.260365994414315, 0.567451399751008, 0.0361954991240054, 0.819706920534372, 0.278743197210133, 0.372681614942849, 0.404339884407818, 0.31570374686271, 0.0715611553750932, 0.308326343772933, 0.971401036949828, 0.896155826747417, 0.842471160693094, 0.183278617681935, 0.115669144550338, 0.692570134298876, 0.984840789111331, 0.531812484143302, 0.305397884454578, 0.742137187626213, 0.970885640475899, 0.452866520499811, 0.0507103211712092, 0.060659340582788, 0.703413825714961, 0.379121451638639, 0.612184786470607, 0.278446133714169, 0.727485248586163, 0.687288482673466, 0.691161301685497, 0.285216487245634, 0.977690197527409, 0.0843506278470159, 0.238997458014637, 0.947240740060806, 0.964063014602289, 0.199624139815569, 0.150643441360444, 0.899659895338118, 0.650838076137006, 0.759135586675256, 0.97109033470042, 0.671637673862278, 0.910862527554855, 0.309689442394301, 0.290938223013654, 0.864073792705312, 0.061465117149055, 0.805029795505106, 0.403617419535294, 0.572206981247291, 0.251219276338816, 0.837345013860613, 0.410749061964452, 0.204638465307653, 0.558684003772214, 0.787793083814904, 0.0707918375264853, 0.880248227855191, 0.833252473734319, 0.804981841007248, 0.0996369901113212, 0.439199006650597, 0.456200936343521, 0.564606475178152, 0.418469196651131, 0.81310557294637, 0.697536255698651, 0.393647420918569, 0.794448639731854, 0.884066861122847, 0.63048969488591, 0.104139920556918, 0.368109745671973, 0.218577959109098, 0.472161548212171, 0.191892711212859, 0.328668451169506, 0.163859887048602, 0.607273950008675, 0.291953943669796, 0.160636848304421, 0.656851490028203, 0.652644066605717, 0.501443649875, 0.395864789839834, 0.307785919401795, 0.838957449886948, 0.018892046995461, 0.871499059721828, 0.315569196129218, 0.164971192367375, 0.963719450170174, 0.697290129493922, 0.179318260867149, 0.180716702248901, 0.068906630622223, 0.0100676519796252, 0.665894247824326, 0.177837731549516, 0.654883876908571, 0.55364502989687, 0.426120359217748, 0.617224598303437, 0.467295951908454, 0.482179872691631, 0.442038353532553, 0.693840709282085, 0.334375711157918, 0.726919435895979, 0.895768203539774, 0.906833593500778, 0.940631975885481, 0.613184029003605, 0.08353744703345, 0.863822285784408, 0.709816799033433, 0.393898547394201, 0.913594795856625, 0.0264969456475228, 0.218398814555258, 0.942084203939885, 0.21820614230819, 0.461692295502871, 0.657360009849072, 0.539682386443019, 0.518321029143408, 0.344798678532243, 0.0525298067368567, 0.867604600964114, 0.828277943655849, 0.784099558135495, 0.254694419214502, 0.00818910845555365, 0.245160665130243, 0.624527198961005, 0.97616670653224, 0.0238865939900279, 0.188539544353262, 0.680208658799529, 0.476698282873258, 0.00241417717188597, 0.011630172142759, 0.794553727377206, 0.0938743620645255, 0.659514752915129, 0.165272709680721, 0.0462757842615247, 0.743563664145768, 0.284325551707298, 0.839794972678646, 0.00638800882734358, 0.0426358177792281, 0.809704408515245, 0.668590959860012, 0.729849186027423, 0.322763935429975, 0.557412728434429, 0.0780136624816805, 0.0180526173207909, 0.538060774328187, 0.151493744459003, 0.204281584825367, 0.990577135235071, 0.0364759762305766, 0.697561708977446, 0.13037530798465, 0.24913707212545, 0.100820435211062, 0.956552741583437, 0.405560777988285, 0.782859659055248, 0.148791774176061, 0.84037228138186, 0.292043005116284, 0.486753593664616, 0.793198736850172, 0.585643051424995, 0.0103764631785452, 0.737066522007808, 0.539806050481275, 0.27011151611805, 0.544949143892154, 0.81104829069227, 0.908704868284985, 0.59090543910861, 0.615740025416017, 0.0210923014674336, 0.206986459437758, 0.164835751987994, 0.885489489184693, 0.365558996796608, 0.730518119875342, 0.977993272477761, 0.104618337005377, 0.118507224367931, 0.10304336133413, 0.407085277140141, 0.813283570343629, 0.839348285226151, 0.875267639057711, 0.89560622558929, 0.811829747632146, 0.0961165758781135, 0.712068319320679, 0.493791820947081, 0.334935765247792, 0.598888869863003, 0.0935703441500664, 0.575455296086147, 0.214096240000799, 0.0607975674793124, 0.843597997212783, 0.00634967186488211, 0.727124243276194, 0.700389688834548, 0.535140583757311, 0.729351684218273, 0.560636835871264, 0.827403170755133, 0.461330398451537, 0.189424587180838, 0.19697798974812, 0.374801269965246, 0.815569592406973, 0.623641214100644, 0.836007058853284, 0.206393235595897, 0.33163303299807, 0.285156527068466, 0.534861093387008, 0.308685541618615, 0.288575270678848, 0.683911980362609, 0.315246080746874, 0.686437104362994, 0.998082282952964, 0.687769313808531, 0.430439732503146, 0.696036300156265, 0.387029045261443, 0.359926207223907, 0.587543659145012, 0.794610844692215, 0.42743046884425, 0.0773863452486694, 0.719328455626965, 0.482017616042867, 0.175667126430199, 0.835850498871878, 0.554089032579213, 0.712603962980211, 0.27217765734531, 0.419105790788308, 0.915722148492932, 0.232817524345592, 0.843577168649063, 0.258882690453902, 0.614653882803395, 0.877081350889057, 0.757483114721254, 0.882197198458016, 0.534815192921087, 0.465004343306646, 0.442253352841362, 0.150957140140235, 0.447470638900995, 0.000298249768093228, 0.632518547819927, 0.825685282004997, 0.317404573317617, 0.782694867113605, 0.570090850349516, 0.594193807337433, 0.20209263660945, 0.418814935022965, 0.149410914629698, 0.957515426445752, 0.608834633138031, 0.793488681316376, 0.508922564564273, 0.0693707645405084, 0.459890485974029, 0.28043213672936, 0.848423069808632, 0.323220430174842, 0.188136948505417, 0.900339250452816, 0.379477157490328, 0.0598273959476501, 0.77415093081072, 0.914959042100236, 0.5234616023954, 0.856366596417502, 0.393946098396555, 0.951558733824641, 0.561738717602566, 0.503711409401149, 0.303696513175964, 0.625382698141038, 0.44385007978417, 0.526141573209316, 0.857087143929675, 0.40474322787486, 0.617691269377246, 0.667835153173655, 0.497402461012825, 0.867945380043238, 0.717452805954963, 0.318543582689017, 0.825374948326498, 0.435791433090344, 0.332560933660716, 0.251502530649304, 0.0193231755401939, 0.960704421857372, 0.0960206014569849, 0.414551368914545, 0.769514151616022, 0.969297474948689, 0.523372775875032, 0.839491467224434, 0.828399189282209, 0.538313641911373, 0.456359246280044, 0.561057960614562, 0.654794863425195, 0.26877455599606, 0.23688818491064, 0.211368417367339, 0.0530447692144662, 0.651510132476687, 0.591765332967043, 0.146896480116993, 0.650342544540763, 0.727255260339007, 0.30485067027621, 0.556110511999577, 0.455652617616579, 0.514856203924865, 0.88000097963959, 0.430608039489016, 0.773750788299367, 0.0997510710731149, 0.0902836229652166, 0.733665966195986, 0.350341434823349, 0.0745488649699837, 0.845994964009151, 0.43946144473739, 0.779273167951033, 0.558230826864019, 0.87243202724494, 0.974893487291411, 0.838405642425641, 0.799043581355363, 0.531198410084471, 0.982450593262911, 0.949014110025018, 0.184045800473541, 0.861022509867325, 0.0951836986932904, 0.176963936071843, 0.95178421260789, 0.365311736008152, 0.900549330050126, 0.299034296534956, 0.337875312659889, 0.52557465666905, 0.370671980315819, 0.861705562332645, 0.605602211784571, 0.0721202297136188, 0.122607668861747, 0.721191850956529, 0.25637543015182, 0.670503862900659, 0.892127589788288, 0.347891882993281, 0.0585838158149272, 0.357122286921367, 0.00298785441555083, 0.0722430241294205, 0.0407224406953901, 0.0880281231366098, 0.672511856770143, 0.604069537483156, 0.25281719584018, 0.121225598966703, 0.932715235976502, 0.393521059770137, 0.674706543795764, 0.308656068053097, 0.478326103417203, 0.325748187489808, 0.0340816418174654, 0.472968061920255, 0.223582203499973, 0.0748525753151625, 0.805097041884437, 0.412289752392098, 0.856350753922015, 0.776770748198032, 0.848034265683964, 0.991151033667848, 0.459523999365047, 0.861050434643403, 0.165547493612394, 0.0669261033181101, 0.507747764000669, 0.650763130513951, 0.567126953741536, 0.242238213540986, 0.446969089563936, 0.237529876409099, 0.106129030929878, 0.275931604206562, 0.0285364496521652, 0.0372736367862672, 0.0850941939279437, 0.0842146496288478, 0.438625522656366, 0.315954609774053, 0.387999935541302, 0.892107474151999, 0.0465112645179033, 0.955590133322403, 0.921702796360478, 0.409322163555771, 0.627664999105036, 0.63828356936574, 0.89512117090635, 0.9337636413984, 0.778690489474684, 0.350357434013858, 0.0815830968786031, 0.0615923248697072, 0.323831832036376, 0.846981924260035, 0.167217771057039, 0.633416843833402, 0.257420306326821, 0.672507885843515, 0.0398664611857384, 0.34205933380872, 0.111982707632706, 0.902539143804461, 0.789039385039359, 0.286536075407639, 0.812689174897969, 0.431016342248768, 0.958897166652605, 0.0899108066223562, 0.885546241654083, 0.163874950027093, 0.953977546654642, 0.43444281606935, 0.10066967504099, 0.307086375309154, 0.4949071474839, 0.479838283034042, 0.981689392589033, 0.764636376872659, 0.05384885892272, 0.11222327244468, 0.333402450662106, 0.160473884548992, 0.308172315359116, 0.760786216473207, 0.545348669169471, 0.150682647712529, 0.537622444797307, 0.406551869353279, 0.260519229574129, 0.5959738639649, 0.249715843936428, 0.318524213973433, 0.356031855102628, 0.246559896273538, 0.709287067875266, 0.687685740878806, 0.0333801412489265, 0.503032173262909, 0.598019993864, 0.725675330730155, 0.532071472844109, 0.379402260994539, 0.154071628581733, 0.53729833336547, 0.117825147230178, 0.2165211297106, 0.329415601212531, 0.873418217059225, 0.0110098842997104, 0.0306536257266998, 0.139327592216432, 0.790175680536777, 0.71671792306006, 0.462151373969391, 0.898199592251331, 0.500299970852211, 0.112697812030092, 0.906434047967196, 0.47580277826637, 0.493016127264127, 0.911975855240598, 0.274585980456322, 0.466096302960068, 0.429454576224089, 0.640972981229424, 0.601657977094874, 0.653447618475184, 0.209610993275419, 0.0414472648408264, 0.162412288831547, 0.112214140361175, 0.256625432055444, 0.150176780764014, 0.843788048718125, 0.275565811432898, 0.369379961397499, 0.170202710665762, 0.697747039841488, 0.752361614257097, 0.253324242774397, 0.199850754113868, 0.613051416352391, 0.14895365270786, 0.207936158869416, 0.634348075371236, 0.139885020442307, 0.797255595680326, 0.717169197043404, 0.363003350794315, 0.0289896454196423, 0.922463394934312, 0.726825435645878, 0.279703340260312, 0.581307513639331, 0.0910566870588809, 0.695524317445233, 0.136323905782774, 0.273677270160988, 0.438514165580273, 0.612322607077658, 0.101531272754073, 0.957795267691836, 0.743686625501141, 0.434636372607201, 0.84360718447715, 0.907005136832595, 0.567018595291302, 0.374701619613916, 0.95118763926439, 0.0101108227390796, 0.622540534008294, 0.894813724327832, 0.194176353514194, 0.643112786114216, 0.0329897392075509, 0.0412496372591704, 0.610606368165463, 0.755234421463683, 0.167352236574516, 0.171399626182392, 0.208152057603002, 0.16639208025299, 0.637399424798787, 0.740623180987313, 0.211824713507667, 0.276416439795867, 0.0843201426323503, 0.970507928868756, 0.16904612374492, 0.888202630681917, 0.350148379569873, 0.312934899237007, 0.348322758451104, 0.305859880521894, 0.578076684847474, 0.616197705967352, 0.959895141189918, 0.420815402176231, 0.0590408069547266, 0.949708157917485, 0.0631893062964082, 0.276434261351824, 0.630146943731233, 0.914313392480835, 0.0607790136709809, 0.700883653014898, 0.129254860104993, 0.737124798819423, 0.564381410833448, 0.665663565276191, 0.0753041738644242, 0.790271545760334, 0.654744411585853, 0.0947543354704976, 0.0582317048683763, 0.98610974685289, 0.960545574314892, 0.841464977711439, 0.728372962679714, 0.275766567559913, 0.39773584343493, 0.761389432940632, 0.180369487265125, 0.0564974048174918, 0.27977902838029, 0.0947582663502544, 0.125607468187809, 0.639084407826886, 0.328990834765136, 0.236750792711973, 0.920869629597291, 0.649179069092497, 0.631485994439572, 0.847200066084042, 0.851502482313663, 0.993434195639566, 0.916794412303716, 0.252669917885214, 0.662314370740205, 0.434480418916792, 0.769580998225138, 0.678719060262665, 0.314405072713271, 0.561774543253705, 0.870040198788047, 0.63610822102055, 0.587359967408702, 0.678053110605106, 0.0260184416547418, 0.17076116008684, 0.935897259740159, 0.761679488699883, 0.835001763654873, 0.0702237493824214, 0.743579220492393, 0.572012072661892, 0.629740683594719, 0.0440595678519458, 0.862819580826908, 0.856003381777555, 0.241303270217031, 0.0600099607836455, 0.598675607470796, 0.671177977928892, 0.786149638239294, 0.886508960044011, 0.0613691885955632, 0.117764703696594, 0.577109514269978, 0.201753670815378, 0.128645387245342, 0.459685212932527, 0.0725816420745105, 0.194702633423731, 0.755264435196295, 0.917691721348092, 0.291708731558174, 0.706577319651842, 0.863641580566764, 0.70654361974448, 0.889008610742167, 0.132029943168163, 0.165802664356306, 0.0664745620451868, 0.156889059115201, 0.037851284025237, 0.789433171506971, 0.795137293636799, 0.088363723596558, 0.608099054079503, 0.255297139054164, 0.550816427217796, 0.229976368602365, 0.621090648695827, 0.694763591513038, 0.405215688748285, 0.480616413289681, 0.750890844035894, 0.344094151630998, 0.101211828878149, 0.623130171559751, 0.481781566282734, 0.899072993081063, 0.218937191180885, 0.409001502441242, 0.458224921720102, 0.850364141399041, 0.371845358517021, 0.0457586487755179, 0.893691569799557, 0.749893101863563, 0.853905899217352, 0.910591086372733, 0.0807889157440513, 0.956016569631174, 0.00673525780439377, 0.558377490611747, 0.352212990634143, 0.28602178581059, 0.264434565789998, 0.684076786972582, 0.177132337354124, 0.647665737662464, 0.509611174464226, 0.113911070395261, 0.793692895909771, 0.457106937887147, 0.83496509026736, 0.991026122821495, 0.45775284175761, 0.951118225231767, 0.260433063842356, 0.412081338930875, 0.398526948876679, 0.878461124841124, 0.402619809145108, 0.412865416379645, 0.954286998137832, 0.974175977753475, 0.083318036980927, 0.531442848965526, 0.50139123480767, 0.312912008957937, 0.622782765422016, 0.388235844438896, 0.533883999567479, 0.473579135490581, 0.00124250934459269, 0.568553478457034, 0.380940787959844, 0.154865806689486, 0.992089909967035, 0.636303316568956, 0.658234864706174, 0.45976790599525, 0.818449952406809, 0.514354998711497, 0.644257436972111, 0.42699042125605, 0.497820805991068, 0.195978293195367, 0.746059704571962, 0.761596590513363, 0.7089781423565, 0.591278779320419, 0.303694584639743, 0.502770202700049, 0.442294056992978, 0.599945121677592, 0.806293145753443, 0.0215481813065708, 0.557498360751197, 0.874574055895209, 0.338767295936123, 0.825636050431058, 0.377457735594362, 0.425295381806791, 0.524021290242672, 0.459253369597718, 0.834261526586488, 0.154065145878121, 0.185625352663919, 0.275423386832699, 0.847513092681766, 0.0162283102981746, 0.932800639886409, 0.603662441950291, 0.549734102794901, 0.0837756420951337, 0.138990873238072, 0.916312876855955, 0.232130161020905, 0.654345789924264, 0.556627693818882, 0.845363624161109, 0.97336501465179, 0.713808447355404, 0.909822562709451, 0.0319200272206217, 0.23810705775395, 0.174503847956657, 0.627220463007689, 0.127779192524031, 0.329843790736049, 0.208434893051162, 0.437645294936374, 0.31375979189761, 0.734344246564433, 0.970942810876295, 0.438999654259533, 0.129103626124561, 0.793507620459422, 0.857977051986381, 0.67953901970759, 0.382742309942842, 0.644177645212039, 0.207262916490436, 0.456010397756472, 0.840740320039913, 0.780265301465988, 0.622139233164489, 0.363952399929985, 0.469467108836398, 0.881086404202506), actual = c(0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.553545259172097, 0.427066240459681, 0.928924408508465, 0.879183722194284, 0.372813975904137, 0.0849204547703266, 0.185200048377737, 0.248517516301945, 0.972205366240814, 0.175937860272825, 0.29740926483646, 0.177856435067952, 0.54987583309412, 0.926613162504509, 0.623376972973347, 0.995663726469502, 0.0694905053824186, 0.883967674337327, 0.100525211775675, 0.16361204860732, 0.0322660950478166, 0.188483601436019, 0.482381742214784, 0.34511242271401, 0.108717137249187, 0.0858348405454308, 0.119930771179497, 0.681731098797172, 0.274988326709718, 0.266401896486059, 0.933536902070045, 0.0823281980119646, 0.796490027103573, 0.786480345763266, 0.885140919592232, 0.469166294205934, 0.5330254600849, 0.0935750645585358, 0.919401035178453, 0.462205442134291, 0.185791645664722, 0.379412932321429, 0.661133252549917, 0.216575714992359, 0.66335735260509, 0.324423821642995, 0.751940541900694, 0.792235780041665, 0.802989500574768, 0.970741912256926, 0.604920069919899, 0.998510781675577, 0.323315804358572, 0.250296991085634, 0.576650936855003, 0.625906931702048, 0.720699115190655, 0.298825138714164, 0.79749919148162, 0.175754863768816, 0.180132172070444, 0.384371358435601, 0.421485849423334, 0.0191708931233734, 0.226989255053923, 0.682258798973635, 0.465152814751491, 0.915489014936611, 0.738574749324471, 0.157227998366579, 0.129364535911009, 0.620678732171655, 0.00769835826940835, 0.646405516192317, 0.0207766427192837, 0.678867493290454, 0.193675739923492, 0.86962575931102, 0.663124195998535, 0.435731106204912, 0.227465422824025, 0.086377008119598, 0.477468098048121, 0.067721962928772, 0.253991047618911, 0.25930565665476, 0.479382940102369, 0.646845594514161, 0.783757312456146, 0.864591051824391, 0.995197806274518, 0.853795324452221, 0.751615410670638, 0.998073264490813, 0.397997858934104, 0.365965174743906, 0.652265961049125, 0.271112349350005, 0.692928286502138, 0.893678079592064, 0.860804327530786, 0.599785462254658, 0.377051396295428, 0.134076616261154, 0.663567915558815, 0.0863808677531779, 0.486025215825066, 0.0790813139174134, 0.655175419291481, 0.404487154446542, 0.356648030458018, 0.665955220581964, 0.627957408782095, 0.660591531079262, 0.717988968128338, 0.382161184214056, 0.395462305052206, 0.960487502627075, 0.206584273139015, 0.396457844879478, 0.443854105193168, 0.342646634904668, 0.805103315738961, 0.0737944708671421, 0.782874088967219, 0.904258935712278, 0.169439681340009, 0.493608970195055, 0.379336728714406, 0.383747511543334, 0.528265465050936, 0.773300009546801, 0.146164800971746, 0.600231549236923, 0.149402234237641, 0.660344027914107, 0.00584247801452875, 0.249012414366007, 0.702166432747617, 0.754494474967942, 0.0963504409883171, 0.793006703257561, 0.895664525218308, 0.934578447602689, 0.705331772333011, 0.0340771519113332, 0.261333614354953, 0.0849193932954222, 0.17894387873821, 0.791013607522473, 0.771797245601192, 0.829571977024898, 0.468895065132529, 0.723223740700632, 0.247360350564122, 0.488007855601609, 0.283667175564915, 0.939361251424998, 0.740234696073458, 0.314529144670814, 0.675240997225046, 0.216282041976228, 0.76974392356351, 0.662610093830153, 0.0424696705304086, 0.746502556605265, 0.566777846775949, 0.656480361474678, 0.0445085484534502, 0.540418825810775, 0.980545689351857, 0.0328120617195964, 0.675576187903062, 0.45986607321538, 0.0553341552149504, 0.887333458522335, 0.39258004934527, 0.683346933452412, 0.230232290923595, 0.195240244967863, 0.820593252079561, 0.662692913785577, 0.777776585891843, 0.695983804995194, 0.934178882976994, 0.557648301357403, 0.358837377745658, 0.295703847659752, 0.380479692015797, 0.370711916824803, 0.179116667248309, 0.0725438522640616, 0.0724869740661234, 0.872730281669647, 0.772827732376754, 0.539395990315825, 0.652436788426712, 0.158570558065549, 0.207948880968615, 0.14509041630663, 0.807986125349998, 0.67172145540826, 0.263956145849079, 0.937782447086647, 0.00334561779163778, 0.844485703390092, 0.564614901784807, 0.290633372031152, 0.370565895456821, 0.637828023871407, 0.060090892482549, 0.152146164095029, 0.390428726095706, 0.763695415342227, 0.362493098946288, 0.483485132455826, 0.551245950395241, 0.939759196480736, 0.399112488841638, 0.0705508878454566, 0.493953129043803, 0.109589600469917, 0.223448275122792, 0.107885057572275, 0.174438138259575, 0.831522574881092, 0.300347046693787, 0.706227377988398, 0.50453751347959, 0.794862372800708, 0.583104189485312, 0.400306977098808, 0.20756402448751, 0.00870522973127663, 0.538726447150111, 0.71639166562818, 0.0463188332505524, 0.194137474056333, 0.934701539343223, 0.398644766770303, 0.636908177984878, 0.987693075090647, 0.823580997763202, 0.690937484847382, 0.35382606042549, 0.101857269182801, 0.469465762842447, 0.965521147008985, 0.84843117184937, 0.22124887839891, 0.00853542983531952, 0.289425735361874, 0.336112641263753, 0.986148163443431, 0.124783959006891, 0.549410713138059, 0.226829231483862, 0.825822850922123, 0.0446271495893598, 0.305559794651344, 0.101224758895114, 0.448575421702117, 0.64256223780103, 0.192566389450803, 0.165485754376277, 0.755103613948449, 0.848865480162203, 0.607345651369542, 0.255571994232014, 0.487033281475306, 0.790885746246204, 0.32344161439687, 0.510736670810729, 0.0948228996712714, 0.914728599134833, 0.683408474549651, 0.721378671936691, 0.137704513501376, 0.290459539042786, 0.0417890185490251, 0.536207951139659, 0.547580484300852, 0.973055352224037, 0.642704963916913, 0.1676396636758, 0.736127665033564, 0.181476967874914, 0.790320527739823, 0.12022232147865, 0.383694474352524, 0.246393111767247, 0.572379529010504, 0.833130156621337, 0.638584024272859, 0.0776197188533843, 0.996128196595237, 0.706980728777125, 0.568036577431485, 0.363553135888651, 0.916434213984758, 0.725699293892831, 0.712164282333106, 0.858655831310898, 0.898781714262441, 0.669220193754882, 0.0364966511260718, 0.565422062762082, 0.301802121568471, 0.641792802605778, 0.981787952128798, 0.796411119401455, 0.59664148860611, 0.146177383838221, 0.85369089175947, 0.918623377336189, 0.111966421129182, 0.0609689874108881, 0.431586853927001, 0.386563050560653, 0.660791117465124, 0.524732117541134, 0.241991176269948, 0.843865944538265, 0.306397615000606, 0.228859125170857, 0.215055603301153, 0.881771613843739, 0.683880819007754, 0.0410816587973386, 0.260928749805316, 0.342930808663368, 0.520050324965268, 0.269119892502204, 0.535411512246355, 0.411533061880618, 0.0796246519312263, 0.946843543555588, 0.0797457541339099, 0.5749208512716, 0.00520402030088007, 0.224250373896211, 0.560395203530788, 0.0516387524548918, 0.278568609151989, 0.311469735112041, 0.495244204299524, 0.484195492928848, 0.186889220960438, 0.152748365886509, 0.250597535865381, 0.897574662230909, 0.512130320537835, 0.435821001185104, 0.739637524355203, 0.794046869967133, 0.193568160524592, 0.893551046960056, 0.967387403361499, 0.657239457592368, 0.894682416459545, 0.640906595159322, 0.271867278730497, 0.00527004129253328, 0.951285356655717, 0.917926253052428, 0.0670679733157158, 0.763760243542492, 0.396852500271052, 0.838866114616394, 0.542908134870231, 0.63165000313893, 0.0296352654695511, 0.0173918968066573, 0.803010292816907, 0.896002526627854, 0.177122247871011, 0.665275121340528, 0.182537921238691, 0.196945640724152, 0.589700829936191, 0.394729514373466, 0.978881333488971, 0.0463966780807823, 0.113367192912847, 0.480140876956284, 0.153260822407901, 0.820531885605305, 0.337786849355325, 0.0884116077795625, 0.795866968343034, 0.525297760497779, 0.800093045923859, 0.862456595292315, 0.709836730966344, 0.80987973511219, 0.334307628916577, 0.767571336822584, 0.48065905738622, 0.633225724101067, 0.791050713974983, 0.596234701108187, 0.0582957929000258, 0.61075867456384, 0.768110886216164, 0.31621421710588, 0.837783478433266, 0.0462655955925584, 0.776399892522022, 0.854825400048867, 0.211970432894304, 0.522348592989147, 0.963676966261119, 0.169739743461832, 0.0793844719883054, 0.0021359974052757, 0.567720010178164, 0.457976168720052, 0.637622477253899, 0.772982901195064, 0.981904026353732, 0.0828714026138186, 0.489833031548187, 0.415280015673488, 0.237263715360314, 0.676590615184978, 0.647275771014392, 0.512038770364597, 0.591013610363007, 0.564528812421486, 0.474576558219269, 0.344507742207497, 0.504583065630868, 0.697168447310105, 0.343819636153057, 0.114788381382823, 0.151062479941174, 0.807326667942107, 0.267120805568993, 0.459025738760829, 0.555110703455284, 0.551247441442683, 0.361277944641188, 0.58649456454441, 0.0728859431110322, 0.256867192219943, 0.025006259791553, 0.686064208857715, 0.609926169272512, 0.344937847694382, 0.410053865984082, 0.614696306176484, 0.441214939113706, 0.0731069971807301, 0.591338682919741, 0.466417207382619, 0.786203495459631, 0.835904499748722, 0.30082998261787, 0.740940330550075, 0.54596845805645, 0.442942941794172, 0.291866379091516, 0.655663482379168, 0.800977409817278, 0.766388973221183, 0.233353591989726, 0.343457344453782, 0.599200845696032, 0.945629081688821, 0.320624501910061, 0.00739365979097784, 0.183638295158744, 0.268469435628504, 0.826799018541351, 0.0406942514237016, 0.133010215125978, 0.683476110687479, 0.647202377673239, 0.807777640875429, 0.254760057432577, 0.78341522696428, 0.667966632870957, 0.291554231895134, 0.0443926705047488, 0.322100466350093, 0.0378620747942477, 0.295611995039508, 0.473224631510675, 0.0852217422798276, 0.643550432985649, 0.807295008795336, 0.576829642057419, 0.0131223658099771, 0.475659534102306, 0.663408792112023, 0.753879768773913, 0.288856983184814, 0.0935515521559864, 0.424470196245238, 0.468537371838465, 0.877105500316247, 0.234908105805516, 0.499092607526109, 0.0133173777721822, 0.599444567691535, 0.455206367420033, 0.449888400733471, 0.867176150437444, 0.666680482216179, 0.890135693596676, 0.465955244377255, 0.525573688559234, 0.966853292658925, 0.514581692172214, 0.305405578343198, 0.207229850348085, 0.559739596676081, 0.545797704951838, 0.766023377887905, 0.332025773590431, 0.455584480194375, 0.915881591150537, 0.148171477718279, 0.938362780027092, 0.453150180634111, 0.289433512836695, 0.293460801476613, 0.223131530918181, 0.186633002012968, 0.228848614031449, 0.236474501434714, 0.0719225420616567, 0.0602203423622996, 0.67000353266485, 0.399946943856776, 0.663444323698059, 0.991481012199074, 0.0915620706509799, 0.454928402323276, 0.0804485152475536, 0.794331538723782, 0.0138611444272101, 0.0462415141519159, 0.625553419813514, 0.654603580012918, 0.938421306665987, 0.429666347568855, 0.556273070629686, 0.237852458143607, 0.84105811547488, 0.317826714133844, 0.490087607875466, 0.164242163999006, 0.693207378266379, 0.434465820435435, 0.493997155455872, 0.475988742895424, 0.895911170402542, 0.719509546877816, 0.862833969760686, 0.906414902070537, 0.511663178447634, 0.900651833973825, 0.679012440377846, 0.831254484597594, 0.658707528142259, 0.518047102726996, 0.388356153620407, 0.749630572739989, 0.606858674902469, 0.956991590093821, 0.315564831485972, 0.385971565265208, 0.400009036064148, 0.749235467752442, 0.196447172667831, 0.440225434955209, 0.116522552911192, 0.105497299460694, 0.995558498427272, 0.674900257959962, 0.761806648224592, 0.0793765168637037, 0.582047804491594, 0.15098685468547, 0.275793078821152, 0.276815867982805, 0.417043393710628, 0.644543074537069, 0.678715172922239, 0.675400107400492, 0.12022158363834, 0.497286789584905, 0.199950652895495, 0.646873363060877, 0.711067781317979, 0.490116584580392, 0.851994040189311, 0.70178056252189, 0.791551055386662, 0.536082000471652, 0.51515849516727, 0.273141370620579, 0.881294171093032, 0.804821252590045, 0.454208948882297, 0.306378472829238, 0.792151697911322, 0.0183487569447607, 0.548345042159781, 0.830180994933471, 0.395882659591734, 0.520401382120326, 0.254051393130794, 0.840860647149384, 0.177955481223762, 0.844269687542692, 0.920232141856104, 0.882007560925558, 0.82929844292812, 0.809167029103264, 0.622955560684204, 0.088943175971508, 0.885319064604118, 0.162188281305134, 0.661531469086185, 0.831285243853927, 0.509671496925876, 0.874305428238586, 0.0774719640612602, 0.559360420331359, 0.474897787440568, 0.409136356320232, 0.652430559042841, 0.123487829696387, 0.869897272437811, 0.980057771550491, 0.450974299572408, 0.621420880779624, 0.667093714466318, 0.467107122996822, 0.625373846851289, 0.833160747075453, 0.50350876012817, 0.707168731605634, 0.554066562326625, 0.652557073626667, 0.677589549915865, 0.0682950890623033, 0.810797698097304, 0.281269676750526, 0.950967862969264, 0.976591010112315, 0.269983988255262, 0.662947009783238, 0.735107456799597, 0.690755394753069, 0.273225461831316, 0.131725851912051, 0.574550738325343, 0.596527089830488, 0.773092601448298, 0.81578302453272, 0.692042782437056, 0.632027888437733, 0.86103472718969, 0.187541278544813, 0.0672887610271573, 0.848232323070988, 0.999832702567801, 0.239783509168774, 0.733275062870234, 0.114950368180871, 0.982450539479032, 0.584045025752857, 0.222309094853699, 0.0117327652405947, 0.803248411044478, 0.206426677992567, 0.967255589319393, 0.674927057232708, 0.874431656207889, 0.65782846021466, 0.0352014903910458, 0.137858534464613, 0.229265470523387, 0.135649463394657, 0.600738681852818, 0.992971936706454, 0.590783184859902, 0.561228086939082, 0.701153793372214, 0.741394445532933, 0.680863882647827, 0.281272848369554, 0.2874250265304, 0.273402535822242, 0.507458611158654, 0.464147840859368, 0.824138134950772, 0.533922238275409, 0.4384910245426, 0.0366123917046934, 0.238845714135095, 0.262840954819694, 0.148097426863387, 0.154603125527501, 0.192706106230617, 0.240499059902504, 0.320134118199348, 0.253665781812742, 0.904609732329845, 0.643006963422522, 0.125198600115255, 0.300660115433857, 0.74462573369965, 0.203400349942967, 0.538394620642066, 0.672413854394108, 0.777680066414177, 0.820594926830381, 0.63866250612773, 0.454388235928491, 0.843554443214089, 0.339066507061943, 0.178846431430429, 0.722938235383481, 0.820416830014437, 0.96333022788167, 0.730436368612573, 0.566454914864153, 0.299131569219753, 0.410936235450208, 0.871899083722383, 0.922340877819806, 0.0925297813955694, 0.838933990802616, 0.221777624217793, 0.669361121021211, 0.719335468253121, 0.276939630042762, 0.718082280131057, 0.945378521457314, 0.901476789265871, 0.505980031797662, 0.041547573171556, 0.333361983997747, 0.134496486978605, 0.0918558624107391, 0.869864554842934, 0.566048402804881, 0.530665486352518, 0.141675767023116, 0.150100206024945, 0.621389864943922, 0.71150339115411, 0.158167482120916, 0.596097217639908, 0.879783051088452, 0.646820122841746, 0.520419993437827, 0.533663941314444, 0.338245193939656, 0.415260507725179, 0.3670118029695, 0.75403183652088, 0.055197739508003, 0.183514476055279, 0.889354313490912, 0.942933890968561, 0.197812605183572, 0.862823054427281, 0.937781328335404, 0.535551136359572, 0.697006920585409, 0.972494658082724, 0.879528696881607, 0.854798505548388, 0.66524731554091, 0.491388044552878, 0.106611260212958, 0.221237587975338, 0.134148214012384, 0.831138543784618, 0.574859651271254, 0.0978497851174325, 0.460925895022228, 0.107310979627073, 0.341059713857248, 0.558058413444087, 0.816278757760301, 0.553815538296476, 0.921920462977141, 0.791485169669613, 0.223766749724746, 0.845891093602404, 0.423667005961761, 0.743595196399838, 0.762215823866427, 0.455404829233885, 0.832104906672612, 0.486732943914831, 0.843880454078317, 0.117426128359511, 0.734385212417692, 0.889724198030308, 0.0418511922471225, 0.130921851610765, 0.67874917970039, 0.852334621129557, 0.463503062957898, 0.54460377804935, 0.164399713743478, 0.0385460194665939, 0.266010183142498, 0.808270634617656, 0.475916020106524, 0.222332353238016, 0.53919375105761, 0.516788524575531, 0.806993245845661, 0.310844893101603, 0.574814067222178, 0.528616666560993, 0.582156032323837, 0.313006992451847, 0.0810491335578263, 0.0384072638116777, 0.811420339625329, 0.982730035902932, 0.00412034546025097, 0.10676576057449, 0.766777546377853, 0.855742876185104, 0.655398649163544, 0.168697108980268, 0.127296761143953, 0.987176733091474, 0.560643086908385, 0.189198084874079, 0.0625575953163207, 0.671661683125421, 0.389910425292328, 0.530542206019163, 0.689909224165604, 0.830408770823851, 0.797011273913085, 0.419954244745895, 0.751368609722704, 0.815764416474849, 0.47300341585651, 0.884758167667314, 0.274057551985607, 0.0988088026642799, 0.743669525720179, 0.279183849226683, 0.561277505476028, 0.250417074654251, 0.685354518936947, 0.97797238943167, 0.375502839218825, 0.496240481035784, 0.790643074316904, 0.597057834966108, 0.922924451995641, 0.244192466605455, 0.918558272533119, 0.637566981138662, 0.477290223818272, 0.8819191432558, 0.548778165597469, 0.581008019158617, 0.00831108167767525, 0.305871309246868, 0.712840780150145, 0.858828793978319, 0.42687801644206, 0.411048624897376, 0.672746668104082, 0.286987874424085, 0.837262193206698, 0.262753044255078, 0.439434518106282, 0.00974732334725559, 0.263382657431066, 0.236171294935048, 0.321568344254047, 0.133400585269555, 0.651631608605385, 0.564940048847347, 0.102179647888988, 0.975021351128817, 0.65347802801989, 0.624559447867796, 0.769746328238398, 0.144400312798098, 0.768873966298997, 0.694914727238938, 0.855115298647434, 0.374508559238166, 0.661874566692859, 0.220214860513806, 0.53567821579054, 0.355370281729847, 0.923474499722943, 0.505581936100498, 0.869511716766283, 0.490070080384612, 0.799502877285704, 0.00443526008166373, 0.228697996586561, 0.31369121838361, 0.67457093950361, 0.311469286680222, 0.964283988345414, 0.694932176265866, 0.0564048681408167, 0.534896573284641, 0.0633229992818087, 0.0971101317554712, 0.0694234683178365, 0.051673177164048, 0.399415434803814, 0.82612094655633, 0.0648626280017197, 0.983990625943989, 0.0248644985258579, 0.921088476898149, 0.469116450985894, 0.943046307191253, 0.687110499246046, 0.380366029217839, 0.176062650280073, 0.323813189053908, 0.213975329417735, 0.745034713065252, 0.0422551890369505, 0.655323869781569, 0.267900235718116, 0.663694423856214, 0.432447683531791, 0.838603786658496, 0.736902639968321, 0.535798569442704, 0.651000478304923, 0.66343559930101, 0.143321895506233, 0.502546613104641, 0.327923309989274, 0.468090586364269, 0.38667313172482, 0.178138460032642, 0.49862194317393, 0.375412163557485, 0.116948760813102, 0.0440025762654841, 0.350163790630177, 0.428598147584125, 0.093810060294345, 0.890044379979372, 0.608290088130161, 0.617687593912706, 0.932591228978708, 0.583395151188597, 0.464269847143441, 0.112032221164554, 0.907358907628804, 0.119604862527922, 0.61091168387793, 0.0177969846408814, 0.611398790031672, 0.140089189168066, 0.468847352778539, 0.445550669683143, 0.540533750317991, 0.123228152515367, 0.279595459578559, 0.280205950839445, 0.978966191178188, 0.0325727234594524, 0.96733574103564, 0.197585161775351, 0.515981563134119, 0.489137480035424, 0.571520657744259, 0.297762417467311, 0.0792570151388645, 0.42842457187362, 0.386258005397394, 0.380782197928056, 0.758228114806116, 0.187241648091003, 0.334191594040021, 0.106257487321272, 0.5509822757449, 0.556651244172826, 0.0913719355594367, 0.17470062407665, 0.836908537428826, 0.991607272764668), actual = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"