Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.3253604        0.3253604 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4620065        0.4646969        0.4680802 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.863604656886309, 0.667788387509063, 0.679073269013315, 0.954806415364146, 0.759989866986871, 0.509477716404945, 0.803504324518144, 0.836696133250371, 0.528004667954519, 0.264138323953375, 0.883351218886673, 0.540376045973971, 0.127050331560895, 0.444660911802202, 0.0338830368127674, 0.286604078486562, 0.0610687339212745, 0.822632954688743, 0.0222817440517247, 0.312017106451094, 0.0830104085616767, 0.694492884213105, 0.543777971994132, 0.104291310766712, 0.166668683756143, 0.510068810777739, 0.820340626407415, 0.566478221677244, 0.992647832725197, 0.283384113572538, 0.469276602612808, 0.706746750511229, 0.457957837730646, 0.733426087303087, 0.745498405303806, 0.276906893588603, 0.625735922949389, 0.666829854249954, 0.262956162914634, 0.795803352957591, 0.0932334391400218, 0.916381233604625, 0.264384078094736, 0.292063379427418, 0.993249149061739, 0.168894381262362, 0.949242956936359, 0.247587514575571, 0.278341789729893, 0.476873155450448, 0.500496751628816, 0.135381576837972, 0.228349348763004, 0.1009464149829, 0.104645690647885, 0.0396975597832352, 0.77428826270625, 0.784216603497043, 0.137402656488121, 0.468847267329693, 0.484356456669047, 0.463398880325258, 0.862030476797372, 0.464825201546773, 0.268616159679368, 0.268211910268292, 0.153084080899134, 0.342194103403017, 0.239043975016102, 0.223562499973923, 0.898248863173649, 0.398279552580789, 0.951990209287032, 0.911012986674905, 0.455757716670632, 0.806640145368874, 0.448132703546435, 0.213773319730535, 0.0392104981001467, 0.858045716071501, 0.596576536539942, 0.305438638897613, 0.133390478556976, 0.0489593998063356, 0.352254479425028, 0.13184270914644, 0.0422773647587746, 0.743875011568889, 0.575424361042678, 0.991851128172129, 0.426366704981774, 0.631749034160748, 0.948385764844716, 0.193075111834332, 0.396359558915719, 0.00199750531464815, 0.768736033234745, 0.268851028522477, 0.324080103076994, 0.679542661877349, 0.0483274338766932, 0.839011886855587, 0.104608311550692, 0.497769642155617, 0.11697010230273, 0.261846842477098, 0.764421658357605, 0.323809382272884, 0.821681032655761, 0.0922052217647433, 0.680203359341249, 0.859129802323878, 0.261907959822565, 0.551810673903674, 0.078175111208111, 0.220279099186882, 0.0844956254586577, 0.924774368060753, 0.477468979079276, 0.403915029251948, 0.278047352330759, 0.214108971180394, 0.698022529482841, 0.196060057729483, 0.0475654744077474, 0.844321229960769, 0.372896644985303, 0.445786772063002, 0.119610430905595, 0.942027288721874, 0.976851142710075, 0.402016714448109, 0.619071912951767, 0.116734712384641, 0.0620525255799294, 0.780108208535239, 0.933997430838645, 0.994164703413844, 0.302353357197717, 0.997040434274822, 0.529126635054126, 0.186432475224137, 0.589190574828535, 0.751374489860609, 0.435863760998473, 0.38923434750177, 0.708584013395011, 0.139444375410676, 0.0457168519496918, 0.148485248442739, 0.922634441405535, 0.598223652457818, 0.614395797252655, 0.459299666341394, 0.957324130926281, 0.241299716057256, 0.931713822297752, 0.821224262705073, 0.28935940656811, 0.95667193015106, 0.411225449526682, 0.266411695396528, 0.0360328182578087, 0.782417175592855, 0.636425527511165, 0.372494927840307, 0.110236720880494, 0.444959849817678, 0.618840387789533, 0.637342721922323, 0.409392128000036, 0.133353390730917, 0.507043751422316, 0.850719693116844, 0.63491384126246, 0.682610716437921, 0.470797859830782, 0.230557689676061, 0.263760989299044, 0.566390756284818, 0.35008994396776, 0.153810445452109, 0.0625661173835397, 0.127446833299473, 0.723981448914856, 0.914180523483083, 0.0524308006279171, 0.201430982211605, 0.476427814690396, 0.130663323914632, 0.985616471385583, 0.229136681417003, 0.827542774612084, 0.590192637871951, 0.593176284106448, 0.777392657473683, 0.534456078428775, 0.636471477337182, 0.280739592388272, 0.437474983977154, 0.691319596255198, 0.313033419894055, 0.419222329743207, 0.392258311854675, 0.667160392738879, 0.756395480129868, 0.274838651297614, 0.566998292924836, 0.347439446952194, 0.978203293634579, 0.123002590378746, 0.583673629676923, 0.410248479573056, 0.685468156822026, 0.494794121710584, 0.900038536405191, 0.556163549888879, 0.218468544539064, 0.219580279197544, 0.301213405095041, 0.757228042930365, 0.330747763160616, 0.125187577214092, 0.930374525953084, 0.499681675806642, 0.272482439875603, 0.0705060209147632, 0.374515424715355, 0.206783364061266, 0.752257252112031, 0.463952417485416, 0.457596181193367, 0.291643746430054, 0.168495910707861, 0.746147739002481, 0.456806938396767, 0.775143672246486, 0.002608876908198, 0.92988819279708, 0.731635017553344, 0.301376685267314, 0.558189118746668, 0.515971555141732, 0.237288826378062, 0.20728854672052, 0.950950815342367, 0.094678960274905, 0.449955083662644, 0.912470673909411, 0.579975015250966, 0.576950607122853, 0.901071800850332, 0.225052135996521, 0.177045749966055, 0.544694991316646, 0.470513598993421, 0.897728327196091, 0.91008329205215, 0.113185666501522, 0.83807231974788, 0.57454922166653, 0.0820678346790373, 0.80911040538922, 0.902602831600234, 0.170675758272409, 0.605619761627167, 0.877360313199461, 0.731479501118883, 0.545452639926225, 0.63831663550809, 0.875677813543007, 0.690319236833602, 0.891800008015707, 0.438329002121463, 0.835796528495848, 0.810057460097596, 0.956747397780418, 0.203252515755594, 0.0399891913402826, 0.658431451534852, 0.469732355093583, 0.219624008750543, 0.543429822660983, 0.721517721191049, 0.0437815317418426, 0.25687891850248, 0.222596978070214, 0.20811209292151, 0.953225437551737, 0.561268353601918, 0.136863036779687, 0.316821909276769, 0.961738645797595, 0.76300105266273, 0.313575104344636, 0.710036328528076, 0.145217023789883, 0.947321231011301, 0.38074659043923, 0.838260071119294, 0.723387851612642, 0.0681028545368463, 0.155867186142132, 0.241689865477383, 0.56720555969514, 0.470523890806362, 0.607562048127875, 0.344161593122408, 0.886533552547917, 0.806067736353725, 0.544340108288452, 0.930749137187377, 0.531946959905326, 0.184036630904302, 0.36135160154663, 0.625832392368466, 0.707766784355044, 0.1001491995994, 0.100912071531639, 0.263398263137788, 0.191465279320255, 0.199531562160701, 0.60189428855665, 0.659294816432521, 0.502043961780146, 0.727060894481838, 0.47307061124593, 0.489002661546692, 0.847274472936988, 0.528052673209459, 0.582574781263247, 0.60494463192299, 0.16856841230765, 0.326544726034626, 0.444532103138044, 0.418663798365742, 0.997195650590584, 0.460382137214765, 0.407928275642917, 0.667942728847265, 0.307516140863299, 0.34951005037874, 0.154677486745641, 0.80359882581979, 0.592812702292576, 0.94202646589838, 0.715667526004836, 0.166113626444712, 0.807133187307045, 0.900061696767807, 0.029305585892871, 0.463056295877323, 0.89893987798132, 0.655527914874256, 0.710357691626996, 0.973347721388564, 0.840115739731118, 0.105812520952895, 0.348230488598347, 0.439429403748363, 0.186847694450989, 0.301567975664511, 0.6846433065366, 0.671755205374211, 0.937520909821615, 0.977405693149194, 0.956572863971815, 0.727724629221484, 0.331360573880374, 0.0658459377009422, 0.280256578931585, 0.803975105285645, 0.553169344784692, 0.0345442353282124, 0.592710008146241, 0.132307725492865, 0.452717932406813, 0.432602789718658, 0.427216671174392, 0.557440573349595, 0.123999686911702, 0.745462333317846, 0.474670907482505, 0.9205742708873, 0.878800852922723, 0.745051292702556, 0.931720520602539, 0.122282055672258, 0.213684724410996, 0.355121730826795, 0.431846426334232, 0.124375675339252, 0.616243688389659, 0.421894251834601, 0.68260579649359, 0.430098101496696, 0.650893912650645, 0.610386897809803, 0.370259273331612, 0.900688474299386, 0.0453082432504743, 0.321891747415066, 0.816304681124166, 0.128442997578532, 0.892761767609045, 0.0386029735673219, 0.104048595298082, 0.701273205457255, 0.845529806334525, 0.0596861757803708, 0.960991460829973, 0.832651124335825, 0.666630198480561, 0.523971608141437, 0.150755809387192, 0.638345470419154, 0.916424138937145, 0.392082675127313, 0.23285754211247, 0.6170910873916, 0.897918032016605, 0.274315429152921, 0.790870332391933, 0.525605363771319, 0.316358835436404, 0.0164593097288162, 0.809618771076202, 0.199154885252938, 0.725442953873426, 0.0214781609829515, 0.707959393737838, 0.236435738857836, 0.199333176016808, 0.125250281766057, 0.718527117976919, 0.562454704195261, 0.19513580086641, 0.370221730088815, 0.744918792974204, 0.0402972849551588, 0.461716183228418, 0.547613288508728, 0.349602903006598, 0.626075652893633, 0.191951404558495, 0.214759802911431, 0.852462445618585, 0.459220809862018, 0.424457879271358, 0.423868365120143, 0.251694553997368, 0.451213159132749, 0.20001668902114, 0.42839791555889, 0.502163880737498, 0.395354718202725, 0.383564363466576, 0.0595523312222213, 0.983065645908937, 0.279876222833991, 0.494307971792296, 0.048968818038702, 0.697636466706172, 0.741867370903492, 0.0505098719149828, 0.613328923704103, 0.0269768931902945, 0.529029170284048, 0.933745117392391, 0.780939964577556, 0.357721604872495, 0.974863964132965, 0.411482752067968, 0.163555682869628, 0.315425985492766, 0.421259844908491, 0.853114719036967, 0.0953807116020471, 0.950272020418197, 0.39615118294023, 0.0091810398735106, 0.256122538819909, 0.271117005264387, 0.159384243423119, 0.780137280700728, 0.887772579677403, 0.569471935043111, 0.72892882139422, 0.166522973915562, 0.345095310593024, 0.0932396720163524, 0.459554211236537, 0.123222843743861, 0.0700609725899994, 0.818997957510874, 0.881155899725854, 0.407779568340629, 0.197662164922804, 0.322917611571029, 0.642768009798601, 0.617344474187121, 0.368804720928892, 0.0235401627141982, 0.436939108651131, 0.684390399372205, 0.75275724940002, 0.659797330154106, 0.199366077547893, 0.860081061720848, 0.588423369918019, 0.770958031993359, 0.876693750964478, 0.361021307995543, 0.660601726500317, 0.447631154675037, 0.179567782208323, 0.255586468381807, 0.639434151118621, 0.377235334832221, 0.45788259524852, 0.00512994127348065, 0.290962946368381, 0.719562686979771, 0.997801972087473, 0.111799772828817, 0.950663330964744, 0.108847540337592, 0.0487150941044092, 0.387387370690703, 0.687307197134942, 0.00651124049909413, 0.529629674041644, 0.183537262491882, 0.271243687486276, 0.0928890265058726, 0.487160020740703, 0.075476934434846, 0.226174775045365, 0.967621407704428, 0.109028604114428, 0.138865868095309, 0.91466458607465, 0.613586923107505, 0.189219555817544, 0.345088331727311, 0.187340476782992, 0.582421742379665, 0.0661613959819078, 0.285315855406225, 0.054601309588179, 0.512153895804659, 0.129626925103366, 0.453927773749456, 0.0389822665601969, 0.170488982228562, 0.354401023825631, 0.0850722300820053, 0.863163706613705, 0.707115611759946, 0.552257782313973, 0.367627972969785, 0.152781778480858, 0.881766848498955, 0.310497994534671, 0.0852311698254198, 0.684154821094126, 0.771534005180001, 0.750314185628667, 0.686281957197934, 0.386889013461769, 0.755167636787519, 0.998480978189036, 0.858328264439479, 0.310263487743214, 0.0310823586769402, 0.758186230668798, 0.17117746011354, 0.304567876271904, 0.311106489738449, 0.128255321644247, 0.460714015644044, 0.947021895786747, 0.193496521562338, 0.702896803384647, 0.640868100337684, 0.975472635822371, 0.359996745362878, 0.880186331225559, 0.385041023371741, 0.741825105156749, 0.820020747603849, 0.229174141539261, 0.978322729701176, 0.958816884318367, 0.228603581432253, 0.511481786845252, 0.026271598180756, 0.869888463523239, 0.493881494505331, 0.840995725244284, 0.438069988740608, 0.799592145252973, 0.612680282443762, 0.220129899680614, 0.597755240742117, 0.591502931434661, 0.846829573158175, 0.365669964347035, 0.783304982585832, 0.384458570508286, 0.995112900156528, 0.352740966947749, 0.943039370002225, 0.0790957279969007, 0.280636822106317, 0.103102495195344, 0.766561066964641, 0.38340021437034, 0.184281727997586, 0.475209194235504, 0.506588181015104, 0.909958635922521, 0.787442289991304, 0.118260528892279, 0.535224667051807, 0.0876469402574003, 0.480122190667316, 0.234712498495355, 0.88784384704195, 0.0159249897114933, 0.735834235791117, 0.03876968100667, 0.271528593497351, 0.619164458708838, 0.717623224016279, 0.485539408400655, 0.98389000329189, 0.413596419384703, 0.366409606067464, 0.447576836915687, 0.0689828104805201, 0.672805654350668, 0.28189405798912, 0.097077283076942, 0.111188879236579, 0.347531870473176, 0.830482340417802, 0.129649097332731, 0.601050231605768, 0.0665114838629961, 0.572859176667407, 0.474104039836675, 0.258233727887273, 0.563579125795513, 0.642208950128406, 0.575773967662826, 0.964771216502413, 0.0974633491132408, 0.216460537863895, 0.207938870182261, 0.0930813490413129, 0.175616407534108, 0.260042274836451, 0.174915360752493, 0.428355685668066, 0.933966207318008, 0.835730018792674, 0.178607509005815, 0.505029029911384, 0.718342169420794, 0.0904890343081206, 0.575067170662805, 0.443970947293565, 0.170322522055358, 0.379057988291606, 0.99479451123625, 0.488782071042806, 0.131474085850641, 0.468328206334263, 0.0858098387252539, 0.37438878393732, 0.390005043009296, 0.71299337525852, 0.389373573940247, 0.800242645666003, 0.0600145771168172, 0.171647026203573, 0.982270336011425, 0.933333258843049, 0.0516258263960481, 0.363687982549891, 0.818500775145367, 0.400730398250744, 0.829878337914124, 0.268413777230307, 0.877033843658864, 0.531445261556655, 0.939541155006737, 0.0870255408808589, 0.714872260577977, 0.948533312650397, 0.92996015958488, 0.263057936914265, 0.756556925131008, 0.241330529330298, 0.0418555987998843, 0.809894449776039, 0.504521067487076, 0.374109745491296, 0.224005605559796, 0.911392365582287, 0.823328031226993, 0.852254373254254, 0.656455148477107, 0.504284208407626, 0.386769023956731, 0.3195591988042, 0.309395176358521, 0.23588677844964, 0.167143462458625, 0.876710521988571, 0.905930256005377, 0.0108393852133304, 0.420619948068634, 0.150111133232713, 0.38594910944812, 0.021637707715854, 0.433576718438417, 0.119469071971253, 0.588720283471048, 0.210508608492091, 0.732454953016713, 0.220850941259414, 0.996187311131507, 0.543600568547845, 0.717169536510482, 0.405916150426492, 0.936313720420003, 0.37930642045103, 0.32272582128644, 0.142010151641443, 0.721565547864884, 0.899973445106298, 0.312598370481282, 0.504263346316293, 0.972345030400902, 0.371504068607464, 0.687804365763441, 0.240396592766047, 0.237758307252079, 0.462671041488647, 0.211226285435259, 0.909081263002008, 0.28771791793406, 0.210018246201798, 0.517146696802229, 0.613352758809924, 0.956871421542019, 0.346124736592174, 0.11811697576195, 0.775528815574944, 0.983475872548297, 0.594284978462383, 0.395611173240468, 0.675700353924185, 0.181664991891012, 0.390935974894091, 0.203135967254639, 0.768692539772019, 0.480928836623207, 0.969527605455369, 0.189923986559734, 0.652513046050444, 0.287622401723638, 0.0182176388334483, 0.573864097939804, 0.900332316523418, 0.14602797687985, 0.226422479376197, 0.91954857413657, 0.30414913687855, 0.31474507576786, 0.209918250562623, 0.984106813557446, 0.46771594369784, 0.182618155376986, 0.170690928585827, 0.495139870559797, 0.544210266321898, 0.248737335205078, 0.529767656698823, 0.755374877247959, 0.877357526915148, 0.769751209532842, 0.94937102147378, 0.415418558754027, 0.20320968166925, 0.417657534824684, 0.970117485616356, 0.990936241112649, 0.608971467940137, 0.445027695503086, 0.307880223495886, 0.228457489516586, 0.937375061679631, 0.0242949749808758, 0.242611234076321, 0.307469555642456, 0.918704163981602, 0.590295364847407, 0.0358278430067003, 0.483072086935863, 0.901962713105604, 0.0127624007873237, 0.583767615724355, 0.152763918507844, 0.00455873878672719, 0.714075897820294, 0.940105927875265, 0.264909966383129, 0.922919681994244, 0.570466972654685, 0.261398744769394, 0.17407973180525, 0.221175165381283, 0.969916085479781, 0.600274634314701, 0.828737095464021, 0.182368232868612, 0.855463359039277, 0.231540644774213, 0.623647862579674, 0.384596856543794, 0.136085631325841, 0.0203226921148598, 0.362331747775897, 0.0918497175443918, 0.118360181804746, 0.505189558723941, 0.158612209605053, 0.351324805989861, 0.247019621077925, 0.961689398624003, 0.531850715400651, 0.618850092636421, 0.213707310613245, 0.777533716289327, 0.559504476143047, 0.06106858397834, 0.811228316975757, 0.0726481035817415, 0.176792489597574, 0.0862606056034565, 0.625539490487427, 0.315225912258029, 0.894179573981091, 0.249100715620443, 0.831811773590744, 0.989132622024044, 0.266073546838015, 0.541464305948466, 0.589088753098622, 0.0749164624139667, 0.96913649677299, 0.948942787246779, 0.973192825680599, 0.925797418924049, 0.641445359680802, 0.0548348422162235, 0.0606683962978423, 0.848776407074183, 0.679543586447835, 0.0987539675552398, 0.0932311418000609, 0.369971232954413, 0.766766972839832, 0.962567514739931, 0.838958224514499, 0.120119582628831, 0.525835578329861, 0.692864587064832, 0.361274268245324, 0.435191469267011, 0.677425839705393, 0.128991147503257, 0.444143527187407, 0.77840488916263, 0.725439029047266, 0.368295575957745, 0.0985589430201799, 0.795761195244268, 0.797305549494922, 0.467715651029721, 0.192746551474556, 0.194862230215222, 0.133364700479433, 0.1873555094935, 0.983804459217936, 0.824877541977912, 0.0885122432373464, 0.167167620034888, 0.829314265167341, 0.435160715365782, 0.835212664678693, 0.26853254949674, 0.39521699398756, 0.964994953712448, 0.659157160902396, 0.337306903209537, 0.533215822884813, 0.351921067805961, 0.485869399970397, 0.887236736482009, 0.431573637295514, 0.366259385598823, 0.0352327313739806, 0.256685201544315, 0.589688376756385, 0.22731546824798, 0.435090235201642, 0.0987369120121002, 0.578610601602122, 0.202654321910813, 0.699033960467204, 0.599129936890677, 0.115175025304779, 0.0927075003273785, 0.634097954723984, 0.194811475696042, 0.154360475484282, 0.359430278185755, 0.559848555130884, 0.201563888695091, 0.567588042933494, 0.533317189430818, 0.225099077448249, 0.366768927546218, 0.389788136817515, 0.819157300051302, 0.816983313299716, 0.685253805946559, 0.0548081351444125, 0.996720885159448, 0.134592376416549, 0.716418997384608, 0.6658901325427, 0.837866457412019, 0.697880081133917, 0.346957134548575, 0.522301144897938, 0.743025932461023, 0.0802743658423424, 0.0440260758623481, 0.3086296338588, 0.486894350033253, 0.431058221962303, 0.909544470952824, 0.859057462075725, 0.45982619584538, 0.0571300948504359, 0.331913096830249, 0.429986360948533, 0.916043225908652, 0.480162003077567, 0.0584017001092434, 0.136496080784127, 0.97324112127535, 0.566526201087981, 0.129203959833831, 0.767685974249616, 0.189432650106028, 0.295235038036481, 0.55860375915654, 0.386046637082472, 0.76079708035104, 0.512410265393555, 0.406887593446299, 0.312908594962209, 0.0221478762105107, 0.27508984412998, 0.162184298038483, 0.140609013149515, 0.294026142451912, 0.0561303377617151, 0.104908370180055, 0.0658308919519186, 0.720354359596968, 0.249764475272968, 0.0243631997145712, 0.21491701179184, 0.0795391614083201, 0.0458755621220917, 0.431880033574998, 0.971036479109898, 0.282754833577201, 0.421090824063867, 0.0442365233320743, 0.330518584931269, 0.292310354067013, 0.156745265005156), actual = c(1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.178192518651485, 0.129632366355509, 0.826708814827725, 0.00526339723728597, 0.929803838254884, 0.307144618127495, 0.892176691209897, 0.196448889328167, 0.88942138873972, 0.563227641396224, 0.406046151183546, 0.989450173219666, 0.519707160303369, 0.0627315037418157, 0.383251807419583, 0.152011889498681, 0.843864255351946, 0.639089384581894, 0.18881476810202, 0.0245852852240205, 0.831819829763845, 0.167749547632411, 0.993687160778791, 0.318424290046096, 0.052047562552616, 0.144402384525165, 0.932199012488127, 0.553393523441628, 0.576231463579461, 0.577804140280932, 0.166528949514031, 0.752960004610941, 0.321515039540827, 0.38849201053381, 0.92282411782071, 0.116727567277849, 0.642210896359757, 0.996709551429376, 0.700085905613378, 0.134092653403059, 0.999930241145194, 0.299803907284513, 0.269105020910501, 0.67471614270471, 0.363126961281523, 0.126839366974309, 0.973515240242705, 0.814096474088728, 0.200448335148394, 0.735024477122352, 0.215251456946135, 0.337468547280878, 0.915801079710945, 0.159127676859498, 0.22561951354146, 0.7492261743173, 0.834358680527657, 0.496443125652149, 0.495056721614674, 0.314669378334656, 0.502804369432852, 0.96525970986113, 0.642165476689115, 0.239978002151474, 0.724794558016583, 0.746551749994978, 0.459219175158069, 0.320594546152279, 0.440805030753836, 0.0331335433293134, 0.0994325964711607, 0.333586341002956, 0.940034732455388, 0.77124739298597, 0.985487571451813, 0.252252661623061, 0.94948914507404, 0.805138327181339, 0.615760224871337, 0.130083168158308, 0.147224367363378, 0.858731105923653, 0.503876321483403, 0.243632931029424, 0.735999034717679, 0.187387871090323, 0.512939527630806, 0.478285608813167, 0.411086970707402, 0.851027538068593, 0.321501629427075, 0.739028272684664, 0.308442542562261, 0.311889312230051, 0.0579801190178841, 0.434758543968201, 0.0840595241170377, 0.673698720987886, 0.0327511895447969, 0.319722117390484, 0.490735307568684, 0.625584473600611, 0.9923265427351, 0.141800286248326, 0.684207102283835, 0.43749568797648, 0.604664217913523, 0.64275249093771, 0.15727227111347, 0.491178188240156, 0.821907186182216, 0.691345947328955, 0.445418511051685, 0.819413572084159, 0.334287483943626, 0.18199105700478, 0.327724296832457, 0.696079785469919, 0.657842684071511, 0.0623058739583939, 0.736324798781425, 0.194559691706672, 0.306525223422796, 0.223620805423707, 0.643357057357207, 0.997184081701562, 0.296012468636036, 0.176064167870209, 0.457010342739522, 0.728715264471248, 0.775281321257353, 0.412298341514543, 0.632581395795569, 0.648903311463073, 0.626286881044507, 0.783243713201955, 0.930961453588679, 0.203310325741768, 0.0710556332487613, 0.977627449901775, 0.367632231907919, 0.1936292455066, 0.0907336694654077, 0.482270072912797, 0.816018064273521, 0.763820637250319, 0.887176387477666, 0.0816848066169769, 0.387752309208736, 0.510250573279336, 0.561203365447, 0.272648967569694, 0.268812481081113, 0.248013306874782, 0.211520158918574, 0.854468592209741, 0.941578165860847, 0.516677355859429, 0.920753077371046, 0.532832098659128, 0.370206725085154, 0.51130991964601, 0.716770330676809, 0.751571868546307, 0.0424682409502566, 0.0453937745187432, 0.464328365167603, 0.378432355355471, 0.174578931182623, 0.889870401704684, 0.0969869941473007, 0.777401008876041, 0.770899923751131, 0.808254291303456, 0.40238867001608, 0.498557862825692, 0.401177047286183, 0.145356691675261, 0.0500512281432748, 0.948418641462922, 0.456058646086603, 0.517652639420703, 0.24491372378543, 0.152578932698816, 0.443192076170817, 0.341689490713179, 0.777154982555658, 0.817918436368927, 0.920592150185257, 0.192693136632442, 0.126264336518943, 0.881572273327038, 0.0960039456840605, 0.348805973306298, 0.540035264799371, 0.37902720249258, 0.86852872511372, 0.748659873148426, 0.651109072612599, 0.375780929112807, 0.701934120617807, 0.523516294080764, 0.534261447843164, 0.399056374793872, 0.563161741010845, 0.706940703326836, 0.223716136999428, 0.671939799794927, 0.698039715411142, 0.582944416906685, 0.0444711351301521, 0.168775534490123, 0.682493783999234, 0.796642861329019, 0.380990400910378, 0.601130775408819, 0.537251200294122, 0.111969361547381, 0.0184589675627649, 0.726354777347296, 0.990612626308575, 0.674267160706222, 0.631424342514947, 0.0444097139406949, 0.655053901253268, 0.264579500304535, 0.106680765282363, 0.818946790415794, 0.789041789714247, 0.505958474939689, 0.541734187165275, 0.152100346283987, 0.894123893929645, 0.554324070224538, 0.764797954354435, 0.23139632306993, 0.018055692082271, 0.0366493037436157, 0.317289514699951, 0.0648797906469554, 0.344117586035281, 0.0388759188354015, 0.611298280768096, 0.167499440256506, 0.266262317541987, 0.0461498608347028, 0.360119296703488, 0.172538923565298, 0.141851161140949, 0.549565638881177, 0.17513151280582, 0.942302627954632, 0.123391931410879, 0.871181257534772, 0.43909179372713, 0.119816286955029, 0.0588421600405127, 0.556783247273415, 0.951651888666674, 0.658023535273969, 0.534784752177075, 0.977598649682477, 0.224352684104815, 0.966590201947838, 0.0759427095763385, 0.902887002564967, 0.0459160380996764, 0.562275569653139, 0.150816341163591, 0.726550281746313, 0.160591060295701, 0.971389600541443, 0.557835003361106, 0.164496407145634, 0.353458341443911, 0.787227510008961, 0.265373429981992, 0.942810712149367, 0.264126205816865, 0.114200439769775, 0.309068013215438, 0.295022277859971, 0.429826209554449, 0.969872463494539, 0.448734429432079, 0.926864075474441, 0.949462020071223, 0.987047950038686, 0.199699080083519, 0.411411749664694, 0.532875437987968, 0.316563170636073, 0.17014286457561, 0.108892177697271, 0.523426392814144, 0.120522656943649, 0.410858629504219, 0.27908317069523, 0.723666733596474, 0.802606653189287, 0.905860589817166, 0.514928680378944, 0.148947388865054, 0.134636872680858, 0.187590138288215, 0.452545782551169, 0.851792949484661, 0.817075841361657, 0.15269310772419, 0.0892413172405213, 0.619341651909053, 0.50277853384614, 0.54389063664712, 0.716443270910531, 0.680707615567371, 0.584978377679363, 0.858148615574464, 0.137665006332099, 0.412120920838788, 0.86115582799539, 0.547184987924993, 0.0280595768708736, 0.91742703341879, 0.124175013042986, 0.203369059134275, 0.98461754526943, 0.748855363111943, 0.647770349634811, 0.57093370356597, 0.0296385574620217, 0.563708197092637, 0.365235046250746, 0.69279561447911, 0.93673222954385, 0.770995412487537, 0.893083039904013, 0.487749087158591, 0.153555392753333, 0.394732312997803, 0.28248031064868, 0.375101774930954, 0.0961914611980319, 0.191344166640192, 0.182572926627472, 0.276815948775038, 0.668013005750254, 0.264957634964958, 0.0580508667044342, 0.51145996991545, 0.776203312212601, 0.792686175089329, 0.36987802083604, 0.729069110471755, 0.698136014863849, 0.282648459542543, 0.240280469413847, 0.538411480840296, 0.33070456655696, 0.28145798528567, 0.240489897318184, 0.910136473132297, 0.864936279365793, 0.366102784872055, 0.824340733001009, 0.127748430008069, 0.776927351951599, 0.076296289684251, 0.104274854296818, 0.220571533078328, 0.314008981222287, 0.940386512083933, 0.0544473060872406, 0.322384136263281, 0.456129115307704, 0.102501748828217, 0.0983938314020634, 0.0768437134101987, 0.215272706234828, 0.131641884800047, 0.623542320914567, 0.0349585346411914, 0.540556514170021, 0.974250715924427, 0.682779183844104, 0.96775912027806, 0.0664393394254148, 0.388604703126475, 0.506888635456562, 0.196758931037039, 0.776292051188648, 0.747974850004539, 0.367901007644832, 0.550168064422905, 0.221326178172603, 0.72460095398128, 0.148220136528835, 0.609084535390139, 0.992804134264588, 0.524319209856912, 0.307747841347009, 0.258385904598981, 0.660264048725367, 0.129837718326598, 0.442786451661959, 0.263418977148831, 0.604014789685607, 0.493882921524346, 0.590324074029922, 0.974370133131742, 0.771395158953965, 0.447772227693349, 0.386380643583834, 0.106670612003654, 0.816813824698329, 0.033907474251464, 0.41581354662776, 0.531404392095283, 0.710891688242555, 0.915211716666818, 0.907473966246471, 0.187746411422268, 0.735085462918505, 0.601609109900892, 0.904115536017343, 0.297847962705418, 0.768337321933359, 0.0514311089646071, 0.534073447110131, 0.335094999521971, 0.965973183978349, 0.07995710009709, 0.866374459117651, 0.742047086590901, 0.50826313206926, 0.623965011909604, 0.940795791335404, 0.426747088553384, 0.967810060596094, 0.135222479235381, 0.706086707301438, 0.39599807956256, 0.227910998975858, 0.0770058368798345, 0.351401726715267, 0.898687926586717, 0.84216117975302, 0.49257664103061, 0.626550267916173, 0.984446318121627, 0.543939932249486, 0.643194486387074, 0.83567972201854, 0.107493095099926, 0.330243672709912, 0.00167352426797152, 0.67744639958255, 0.179022915894166, 0.268694190075621, 0.14857729151845, 0.171720374841243, 0.710535419173539, 0.406044950243086, 0.973796202102676, 0.149678767193109, 0.624466754728928, 0.590328649617732, 0.870505973929539, 0.109943527728319, 0.779853623127565, 0.443885006476194, 0.130418108776212, 0.405415932880715, 0.313564826734364, 0.76861965470016, 0.01817886531353, 0.985321632120758, 0.331621448975056, 0.243322697235271, 0.317110026953742, 0.191912445705384, 0.62115300190635, 0.978836152469739, 0.816255209967494, 0.628991411766037, 0.710463588126004, 0.245886477641761, 0.850413151551038, 0.247477274620906, 0.908244375139475, 0.945107794832438, 0.235879426356405, 0.769007388502359, 0.118486376246437, 0.392742226831615, 0.900389298098162, 0.123091191519052, 0.873213681392372, 0.389269156614318, 0.0868943026289344, 0.86241146363318, 0.166734834900126, 0.816783715039492, 0.203999033896253, 0.117793950485066, 0.718378093326464, 0.713202035287395, 0.303737541195005, 0.832598661538213, 0.987671024631709, 0.145384866045788, 0.335502326255664, 0.343895168974996, 0.185437341686338, 0.418099009199068, 0.706697313347831, 0.644238678039983, 0.287414342630655, 0.651977033121511, 0.292688201181591, 0.976921669440344, 0.683517355239019, 0.130734356353059, 0.128707391675562, 0.766895453212783, 0.512926749419421, 0.762168530840427, 0.692824540892616, 0.552963587222621, 0.692012548679486, 0.58881519199349, 0.13273504562676, 0.430229647085071, 0.458393066423014, 0.834404545836151, 0.796033839927986, 0.14046173566021, 0.744734013220295, 0.942395036108792, 0.231372187379748, 0.209908925928175, 0.562906984472647, 0.660625963937491, 0.702685891883448, 0.114495444111526, 0.0230193398892879, 0.370672138174996, 0.711782016092911, 0.474700857419521, 0.865956627996638, 0.779936005827039, 0.428729105507955, 0.751902810763568, 0.420826613204554, 0.060783677501604, 0.905254387296736, 0.636049220105633, 0.256489333463833, 0.782587380381301, 0.56916437856853, 0.921396248042583, 0.502686748281121, 0.943300280021504, 0.365034548798576, 0.814698414644226, 0.325196796096861, 0.0238822577521205, 0.77515709772706, 0.189285739092156, 0.00557142868638039, 0.211748814908788, 0.483796791639179, 0.982080650283024, 0.347451078472659, 0.745921281632036, 0.940511444816366, 0.860034763580188, 0.54758787387982, 0.272550183814019, 0.485294081270695, 0.540770870866254, 0.978636212646961, 0.289131679804996, 0.804938754066825, 0.980607415549457, 0.143046176293865, 0.484597485046834, 0.971576061565429, 0.0771458882372826, 0.0911551369354129, 0.646837060339749, 0.755870099179447, 0.723551633302122, 0.979694137582555, 0.206347278319299, 0.805682739708573, 0.476008427329361, 0.858519721543416, 0.952173402998596, 0.252204855903983, 0.686608174350113, 0.828062638407573, 0.927987565053627, 0.589299809420481, 0.742090446641669, 0.847602616297081, 0.876603229437023, 0.3659795478452, 0.480052799452096, 0.780022909864783, 0.464173247106373, 0.351080126361921, 0.00849918858148158, 0.0642367224209011, 0.647128513781354, 0.0167891262099147, 0.652604018570855, 0.629500517854467, 0.882898483658209, 0.467423702822998, 0.927071252372116, 0.788965133018792, 0.802706183400005, 0.0617480594664812, 0.437266145832837, 0.386852047871798, 0.246161011280492, 0.430673411116004, 0.447317071491852, 0.797295528929681, 0.505506689660251, 0.376232164446265, 0.0125895808450878, 0.339412980247289, 0.0868584737181664, 0.0547465393319726, 0.0570417209528387, 0.494306785752997, 0.715782829094678, 0.190129259135574, 0.975332658505067, 0.0637082785833627, 0.052140690619126, 0.564700011862442, 0.700387804070488, 0.400421834317967, 0.166354641783983, 0.750026650261134, 0.038295497186482, 0.904379582963884, 0.0912663431372494, 0.614392880117521, 0.502162013901398, 0.558589262655005, 0.255377918714657, 0.482663620496169, 0.75855761137791, 0.0463734848890454, 0.0404476698022336, 0.455667439848185, 0.0712962946854532, 0.444100358290598, 0.690941202221438, 0.568013065727428, 0.654313466977328, 0.0498512100894004, 0.324093159288168, 0.0313920327462256, 0.453080839011818, 0.340152869233862, 0.159075429895893, 0.571102539775893, 0.67263164347969, 0.680554963182658, 0.936986384680495, 0.910625645658001, 0.944423403823748, 0.965126384980977, 0.757006353698671, 0.92552916193381, 0.565539132105187, 0.774113039020449, 0.470515281427652, 0.701048783259466, 0.0246108509600163, 0.44630594085902, 0.205732958624139, 0.392133753746748, 0.242518340237439, 0.0485283359885216, 0.150526081211865, 0.244614916387945, 0.318498429842293, 0.692468768451363, 0.913004438625649, 0.580777157563716, 0.952404740964994, 0.993181495228782, 0.759025061270222, 0.645155829610303, 0.258056093007326, 0.662989583797753, 0.480211302638054, 0.577496185200289, 0.663754244102165, 0.178708140738308, 0.191379945958033, 0.799106422346085, 0.491616176906973, 0.321383483940735, 0.997732629301026, 0.135177693096921, 0.968405636725947, 0.226010657381266, 0.405340854777023, 0.0138648657593876, 0.532295907614753, 0.677905921824276, 0.58963051927276, 0.808683145558462, 0.48384253308177, 0.318165028234944, 0.315956231672317, 0.368896858301014, 0.575942063238472, 0.170999796828255, 0.544713142793626, 0.520170157542452, 0.55586833646521, 0.283442264189944, 0.598557074321434, 0.0886544010136276, 0.541993221035227, 0.608843591529876, 0.106067186454311, 0.118380559142679, 0.371909688226879, 0.818427839782089, 0.432704801904038, 0.191515576792881, 0.324457456823438, 0.377939232159406, 0.529034896520898, 0.279228533152491, 0.746297594858333, 0.81568672019057, 0.928811540827155, 0.534024006687105, 0.889188270550221, 0.457247175741941, 0.201768087921664, 0.922231529140845, 0.122522730845958, 0.914682163856924, 0.168484580004588, 0.623197903856635, 0.28161994414404, 0.372708446579054, 0.555787578923628, 0.912020388292149, 0.37627126602456, 0.803827984724194, 0.0915991119109094, 0.404257755726576, 0.973808897659183, 0.61145542608574, 0.820592934265733, 0.406066331546754, 0.302789584267884, 0.670477124163881, 0.538504177704453, 0.26275871717371, 0.785456997109577, 0.0837033232674003, 0.52367565408349, 0.653359507443383, 0.205615638522431, 0.727393952896819, 0.105298056500033, 0.832479687640443, 0.0585602780338377, 0.146029506810009, 0.555103723658249, 0.920096840243787, 0.391938210697845, 0.679503560531884, 0.925120643107221, 0.124720524065197, 0.0752955458592623, 0.575931454310194, 0.897123766830191, 0.0297095603309572, 0.607425045920536, 0.705288238823414, 0.0845964003819972, 0.350742902373895, 0.453422377118841, 0.799428954953328, 0.629739047260955, 0.748600396793336, 0.653418514179066, 0.684633238939568, 0.779954897705466, 0.605577144771814, 0.367568472865969, 0.713726345449686, 0.914671659469604, 0.691202350193635, 0.268124481197447, 0.605417357292026, 0.166941384784877, 0.968512703664601, 0.0598290550988168, 0.360172849847004, 0.342584965052083, 0.889841864584014, 0.550233791349456, 0.905837458791211, 0.991358359344304, 0.231630900874734, 0.369833734119311, 0.360083961160854, 0.426915362942964, 0.735951361479238, 0.269913978874683, 0.471149024553597, 0.728542141383514, 0.540945498738438, 0.494791221106425, 0.0820752885192633, 0.196495504816994, 0.171684511238709, 0.211444856133312, 0.182880508014932, 0.452278387732804, 0.837091983063146, 0.978676683967933, 0.315212716348469, 0.339263560948893, 0.706210634438321, 0.932590196607634, 0.0526025546714664, 0.814960189629346, 0.314786969684064, 0.0901494992431253, 0.795136574422941, 0.212107563856989, 0.912323511205614, 0.784423359436914, 0.475802086759359, 0.843050224939361, 0.987882210873067, 0.191118945367634, 0.937992647988722, 0.150340435793623, 0.958083221921697, 0.140643093036488, 0.183511443668976, 0.887236333219334, 0.828038306208327, 0.85812094528228, 0.535966881318018, 0.151522758649662, 0.683249734574929, 0.963066353695467, 0.0401186109520495, 0.581337994197384, 0.682309483876452, 0.548518789000809, 0.846909189829603, 0.410117481602356, 0.104538200888783, 0.901537137571722, 0.744479738874361, 0.118684396147728, 0.833406791556627, 0.281734737101942, 0.484411807730794, 0.393560813274235, 0.64160779886879, 0.913595395861194, 0.807256731204689, 0.640307195018977, 0.397418528562412, 0.283546200022101, 0.173089881660417, 0.41526156431064, 0.50064330897294, 0.702475369442254, 0.51756175677292, 0.976404763525352, 0.725220021791756, 0.276479748077691, 0.951404515653849, 0.339697525370866, 0.0276768018957227, 0.367507879855111, 0.542577828513458, 0.915143311256543, 0.272033363347873, 0.673209716798738, 0.36035592854023, 0.298965499969199, 0.462107348488644, 0.942752975272015, 0.782403705641627, 0.191865327302366, 0.53398909070529, 0.879492986248806, 0.723871425492689, 0.914969912962988, 0.451000501867384, 0.765082146041095, 0.0564495904836804, 0.218377483310178, 0.89821622450836, 0.657197404187173, 0.762263225624338, 0.424995881039649, 0.0893530175089836, 0.957715114811435, 0.816329586086795, 0.302176938159391, 0.937756303464994, 0.672828509937972, 0.876908348873258, 0.793594700051472, 0.0665231288876384, 0.642735427711159, 0.269333694130182, 0.926428937586024, 0.901100015034899, 0.797611906193197, 0.275930572301149, 0.892819601576775, 0.461284482618794, 0.961318358778954, 0.651872313348576, 0.364655433688313, 0.409031190210953, 0.100174405379221, 0.345719435019419, 0.128116662148386, 0.976667768787593, 0.522419066401199, 0.992273825686425, 0.904806834645569, 0.851214322960004, 0.894239399582148, 0.48050378751941, 0.667877332540229, 0.764797865180299, 0.693707443308085, 0.346861177589744, 0.603657774860039, 0.397941534174606, 0.535530359717086, 0.380529913585633, 0.441824974492192, 0.205299710854888, 0.816550653427839, 0.748357424978167, 0.147486081114039, 0.490901250392199, 0.351795659400523, 0.516408948926255, 0.591067479224876, 0.389507550746202, 0.655219874577597, 0.175403598230332, 0.315319559536874, 0.515419120201841, 0.079107865691185, 0.939657690934837, 0.851504907943308, 0.0294929170049727, 0.621496038977057, 0.885600090958178, 0.875126180471852, 0.979686909355223, 0.652918880572543, 0.372244553407654, 0.402964550303295, 0.579404842341319, 0.737269218545407, 0.568761586677283, 0.773031643824652, 0.683429928962141, 0.740554166957736, 0.931929512647912, 0.331485704286024, 0.709561749361455, 0.0279056446161121, 0.598157607251778, 0.582091673510149, 0.399227428017184, 0.540033063385636, 0.965404323767871, 0.855015537468717, 0.961691997013986, 0.782974886940792), actual = c(0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"