Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>       0.00000000       1.00000000       0.09949131       0.23425490 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>       0.50000000       0.47745899       0.47745899       0.48363755 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.707451365888119, 0.796419076854363, 0.932955816620961, 0.216782858129591, 0.733265869086608, 0.891086681978777, 0.0675697675906122, 0.123584299115464, 0.484706100309268, 0.12180698197335, 0.620815102709457, 0.421062403125688, 0.968105919659138, 0.786677941679955, 0.355727387359366, 0.953067668713629, 0.644368519075215, 0.840275690425187, 0.109170214738697, 0.257514219731092, 0.574100465280935, 0.3576508408878, 0.151969192316756, 0.948892607353628, 0.861921195173636, 0.617563754552975, 0.0974294564221054, 0.213525538798422, 0.394664561608806, 0.954091701423749, 0.962778970832005, 0.819256253307685, 0.427528192754835, 0.464778104564175, 0.463226820575073, 0.085386730497703, 0.778210906777531, 0.323225845815614, 0.233698031166568, 0.0793575674761087, 0.195838045794517, 0.739390809787437, 0.181810614885762, 0.909021235536784, 0.872146634850651, 0.271752221509814, 0.961496629053727, 0.0544929516036063, 0.0225171188358217, 0.77280476083979, 0.780648716492578, 0.961009000893682, 0.936145871179178, 0.380536815850064, 0.91836600494571, 0.881540581583977, 0.646861313143745, 0.566049769055098, 0.322701710741967, 0.250742466887459, 0.818890007678419, 0.0666325392667204, 0.717083243653178, 0.936928855022416, 0.119999727467075, 0.687186648370698, 0.743479324504733, 0.75545406807214, 0.882155266823247, 0.611250950023532, 0.719010971719399, 0.0865845510270447, 0.91306017478928, 0.934191272128373, 0.176419022027403, 0.823451578384265, 0.0743511696346104, 0.256990235531703, 0.342744899680838, 0.156953679164872, 0.036234907573089, 0.0610557897016406, 0.368410396855325, 0.602452456252649, 0.785970732104033, 0.883361088577658, 0.0879887705668807, 0.619444981450215, 0.685073010856286, 0.466612325282767, 0.864964645821601, 0.654132063966244, 0.395378804765642, 0.218752385582775, 0.705854477826506, 0.931264773476869, 0.47334752464667, 0.340476418146864, 0.675938230939209, 0.252905811648816, 0.417040605098009, 0.662171898875386, 0.775132303824648, 0.954612142406404, 0.531277208589017, 0.389259556308389, 0.491800331743434, 0.389066772302613, 0.335365104256198, 0.913772262632847, 0.78434771974571, 0.31059540854767, 0.163911196403205, 0.0816292031668127, 0.769640720449388, 0.224499131320044, 0.578080046223477, 0.69095927849412, 0.239756857044995, 0.654173983260989, 0.610799017827958, 0.483036305289716, 0.617462328169495, 0.179763270309195, 0.614694508025423, 0.0057585455942899, 0.232472522882745, 0.249941022368148, 0.373573932563886, 0.949029781855643, 0.979758536210284, 0.771615066798404, 0.00860539195127785, 0.582962898304686, 0.213401197921485, 0.760075411293656, 0.257106194505468, 0.00676160235889256, 0.102780809160322, 0.694044010015205, 0.658155533252284, 0.758668872993439, 0.827742380322888, 0.115543393185362, 0.480800160206854, 0.516632577171549, 0.281127037014812, 0.501769804162905, 0.648695445852354, 0.835343595594168, 0.689499268308282, 0.62144366861321, 0.942094518104568, 0.287859457079321, 0.337597669335082, 0.153716523898765, 0.681298785610124, 0.73402822483331, 0.70672474289313, 0.829678172711283, 0.468508146237582, 0.0850876995828003, 0.901643343502656, 0.169471707195044, 0.208124595927075, 0.897106796037406, 0.371794045669958, 0.403908777283505, 0.193728314246982, 0.974726496264338, 0.0570224437396973, 0.267413671128452, 0.114581860834733, 0.0536192241124809, 0.628025718033314, 0.865105279022828, 0.60259045730345, 0.935796860605478, 0.746367548592389, 0.223197210114449, 0.349584967363626, 0.265090101864189, 0.00431785779073834, 0.639561716699973, 0.296463921433315, 0.0271831017453223, 0.15978537639603, 0.384820407489315, 0.069324677111581, 0.0388169670477509, 0.905572892632335, 0.458972246618941, 0.927088781725615, 0.338278089417145, 0.198340988950804, 0.925775707932189, 0.887621007161215, 0.229822773486376, 0.838506697909907, 0.0932429586537182, 0.207027022959664, 0.892305674497038, 0.00591975986026227, 0.839188133832067, 0.319695444311947, 0.548688040813431, 0.363030468579382, 0.722271916456521, 0.367396521847695, 0.789882614044473, 0.60428092465736, 0.159713724628091, 0.605733741307631, 0.0768500210251659, 0.650710875634104, 0.942296934081241, 0.0775816044770181, 0.0976256008725613, 0.0631264243274927, 0.257540910504758, 0.715956142172217, 0.622498584212735, 0.158556526759639, 0.546802051365376, 0.614812256535515, 0.00531164719723165, 0.23818369419314, 0.190717021003366, 0.146641787374392, 0.0266007950995117, 0.343013044912368, 0.876239517936483, 0.161818792577833, 0.494247388094664, 0.251417604507878, 0.0963155850768089, 0.561770987464115, 0.810714927036315, 0.936688165646046, 0.86566575942561, 0.160634713480249, 0.0039441897533834, 0.275657066144049, 0.784053966403008, 0.955179382581264, 0.748038210906088, 0.730199094396085, 0.489033402875066, 0.82607366819866, 0.900196878938004, 0.10179208801128, 0.55496841087006, 0.815086422720924, 0.488842817721888, 0.216729881241918, 0.140893687261268, 0.00485682534053922, 0.438928125426173, 0.439736855914816, 0.738940023817122, 0.766671382822096, 0.285305936587974, 0.320908957859501, 0.139285447075963, 0.442751789232716, 0.949596191290766, 0.910002883989364, 0.210938835516572, 0.138763343682513, 0.118751047644764, 0.836203643353656, 0.209815088426694, 0.847243682481349, 0.905095787951723, 0.713682435452938, 0.596143699483946, 0.343702680896968, 0.759926490951329, 0.155196058331057, 0.407957806019112, 0.372933516511694, 0.0195482110138983, 0.644771606428549, 0.891795886680484, 0.775179677177221, 0.145071741659194, 0.227396958507597, 0.854041322832927, 0.0539773304481059, 0.896501093171537, 0.84726665657945, 0.371949798893183, 0.0662853014655411, 0.116619407664984, 0.265756839420646, 0.257761770160869, 0.827133940067142, 0.478233794914559, 0.6139827799052, 0.288531478261575, 0.01224978803657, 0.253324057208374, 0.895277191884816, 0.975677291164175, 0.580301333917305, 0.722787642851472, 0.913160000462085, 0.23991783009842, 0.0551979772280902, 0.279564333846793, 0.0596276272553951, 0.191738254157826, 0.566503467736766, 0.822248209733516, 0.0687172911129892, 0.824997197138146, 0.788037844933569, 0.579919243231416, 0.272133287042379, 0.207710723392665, 0.382638996234164, 0.794212815584615, 0.949862507171929, 0.846368483500555, 0.197958664502949, 0.646103151142597, 0.285014525987208, 0.04380774917081, 0.731439972529188, 0.461558596231043, 0.652091093361378, 0.254406816558912, 0.18267780658789, 0.860096090706065, 0.890525876544416, 0.0544459009543061, 0.0975590057205409, 0.356261776061729, 0.647362317424268, 0.0467949041631073, 0.350090526277199, 0.0807891313452274, 0.0431156801059842, 0.192516780458391, 0.605284250108525, 0.847703967476264, 0.985723895486444, 0.13015608349815, 0.0181646356359124, 0.586771049536765, 0.0628286434803158, 0.630112124839798, 0.548243642551824, 0.499689700547606, 0.505009278655052, 0.347899202024564, 0.897071788553149, 0.940521945478395, 0.957346910610795, 0.761694757966325, 0.260047069285065, 0.837392829358578, 0.909007747191936, 0.954753868049011, 0.549387573963031, 0.683586065657437, 0.616251857951283, 0.451760115800425, 0.812100611859933, 0.37258369801566, 0.450991340214387, 0.882118633715436, 0.277725118910894, 0.614623149856925, 0.0217282113153487, 0.249939373228699, 0.97175411763601, 0.635457754833624, 0.287329681916162, 0.253099425463006, 0.920610364992172, 0.309891695622355, 0.965904402313754, 0.0250023168046027, 0.663261241279542, 0.841414925875142, 0.513731408398598, 0.146073912736028, 0.274142954964191, 0.604089107597247, 0.874521071091294, 0.167778876144439, 0.663026330759749, 0.251475714845583, 0.531688567949459, 0.59430880844593, 0.785423187771812, 0.943127951584756, 0.700479301624, 0.700446673901752, 0.609432203695178, 0.162142051151022, 0.849344580899924, 0.167292507365346, 0.24341967352666, 0.608476050663739, 0.903017392614856, 0.948909075930715, 0.923531878041103, 0.119136882247403, 0.217599125579, 0.757186732953414, 0.471402455354109, 0.0442349596414715, 0.986678270157427, 0.666523709427565, 0.112048537936062, 0.0415075833443552, 0.485884399851784, 0.760426855180413, 0.575417807558551, 0.806235869647935, 0.187213990837336, 0.878308514598757, 0.429587236605585, 0.85024576052092, 0.0185674845706671, 0.349027215968817, 0.773734908318147, 0.86707770312205, 0.196387669304386, 0.681580238742754, 0.047132620587945, 0.614684487925842, 0.8325931946747, 0.81478936620988, 0.805069178342819, 0.777096239849925, 0.367724485229701, 0.267468783771619, 0.956740012625232, 0.146739598363638, 0.622279703384265, 0.301545384107158, 0.679182941792533, 0.103630434488878, 0.812196680577472, 0.764246720820665, 0.597656274680048, 0.625655730720609, 0.685899418313056, 0.508449871325865, 0.98661376000382, 0.91511114500463, 0.649592866422608, 0.140281307976693, 0.934072943404317, 0.651543316198513, 0.938655053731054, 0.723242563894019, 0.110391756519675, 0.633906714618206, 0.0383742153644562, 0.897242610342801, 0.632915486581624, 0.496290640905499, 0.51801854535006, 0.529960351763293, 0.43076745280996, 0.203478758921847, 0.67584817041643, 0.933523961342871, 0.942290541250259, 0.323834631359205, 0.909442706499249, 0.503733710385859, 0.00706904870457947, 0.453031379031017, 0.829030716093257, 0.888687173603103, 0.906681654974818, 0.990820220904425, 0.0959065638016909, 0.133899485459551, 0.48223720677197, 0.476730764145032, 0.611021799268201, 0.551737063797191, 0.143242513760924, 0.759178013540804, 0.391073433216661, 0.0811574852559716, 0.19879904948175, 0.159343221923336, 0.139584540389478, 0.433759474894032, 0.79585513449274, 0.039920176146552, 0.0318815654609352, 0.156886816723272, 0.532417910406366, 0.13005871954374, 0.903185962233692, 0.20598097727634, 0.320232410915196, 0.557042169384658, 0.678454926703125, 0.0668741622939706, 0.0764463269151747, 0.712860399624333, 0.814851965056732, 0.22713774186559, 0.666126672876999, 0.295954825123772, 0.454714210471138, 0.826925256289542, 0.212164004798979, 0.0735150966793299, 0.437049643136561, 0.702129354001954, 0.62529680179432, 0.180422364734113, 0.162025853991508, 0.918532274430618, 0.245032385224476, 0.209883408155292, 0.31671390379779, 0.016216344665736, 0.420551269548014, 0.993940576910973, 0.553206349024549, 0.328457781579345, 0.205685182940215, 0.844889813568443, 0.497209162451327, 0.0441525494679809, 0.289127414114773, 0.0548968682996929, 0.997538775205612, 0.89530370128341, 0.856454677414149, 0.640357824740931, 0.842294245958328, 0.825146016664803, 0.478133856784552, 0.0983657473698258, 0.139436817262322, 0.515629649627954, 0.511847624788061, 0.938299362780526, 0.868589225225151, 0.0874922685325146, 0.83956040837802, 0.946767334360629, 0.788591649616137, 0.557015313534066, 0.307346534682438, 0.0862552113831043, 0.282640633173287, 0.473977519432083, 0.667288913624361, 0.245129751740023, 0.623492840211838, 0.775553248124197, 0.769706884166226, 0.100061099044979, 0.78074868186377, 0.431667709490284, 0.505144781433046, 0.459184592356905, 0.172571441391483, 0.279240517411381, 0.611974043538794, 0.277099510189146, 0.0689939367584884, 0.282215835293755, 0.228846111567691, 0.201735573587939, 0.473397943424061, 0.000609344569966197, 0.0310529083944857, 0.886706620454788, 0.230406896444038, 0.841862636851147, 0.434624806279317, 0.871492343023419, 0.851915574166924, 0.513769863406196, 0.075862493366003, 0.438175938557833, 0.660343569004908, 0.567122738109902, 0.138142774812877, 0.159793204860762, 0.222442138241604, 0.478100322419778, 0.0847464504186064, 0.579881696496159, 0.267838577274233, 0.5362488925457, 0.618214054731652, 0.284896212629974, 0.301451968960464, 0.916041294112802, 0.482866603182629, 0.660010964609683, 0.803308062255383, 0.180417593102902, 0.497876546112821, 0.181771528208628, 0.921690362039953, 0.188788524596021, 0.692214443814009, 0.071822514757514, 0.972297568572685, 0.331968335667625, 0.328991748159751, 0.0913381394930184, 0.00460507115349174, 0.534396462840959, 0.609590727137402, 0.605525125749409, 0.890133470064029, 0.232389088952914, 0.939578184625134, 0.695578690385446, 0.140173227526248, 0.718937553931028, 0.587449733633548, 0.204810639144853, 0.134699039626867, 0.454185071866959, 0.867292357608676, 0.569809464737773, 0.108321701176465, 0.859449517913163, 0.401606298051775, 0.493662820197642, 0.808032898232341, 0.983967913081869, 0.0970403761602938, 0.948052652645856, 0.0105105838738382, 0.856867964379489, 0.42590489028953, 0.674964132253081, 0.00224111182615161, 0.727484900504351, 0.398719134507701, 0.107432563090697, 0.437721671536565, 0.914936661953107, 0.197289020754397, 0.00198243837803602, 0.498886224813759, 0.801913036964834, 0.627635254058987, 0.0514348854776472, 0.795260422863066, 0.480929700657725, 0.247759555466473, 0.75900322617963, 0.865513755939901, 0.226172216702253, 0.452733230544254, 0.549585599685088, 0.802465964807197, 0.0536166317760944, 0.356557677267119, 0.749670090852305, 0.883909354452044, 0.784018993610516, 0.47676772181876, 0.55101097910665, 0.278399760369211, 0.134593426482752, 0.571284745587036, 0.430935040814802, 0.856113481568173, 0.509982729330659, 0.166234888369218, 0.340054731816053, 0.903026652289554, 0.74050757731311, 0.848069708561525, 0.254113139817491, 0.4370568448212, 0.886694909306243, 0.163795773871243, 0.18137277290225, 0.285309514962137, 0.809743601363152, 0.0364883551374078, 0.148720003431663, 0.678294710582122, 0.203638231148943, 0.874071713071316, 0.00636256951838732, 0.362283144844696, 0.886721522780135, 0.241849835263565, 0.684171103406698, 0.8390662309248, 0.295855324249715, 0.105500851059332, 0.956188871758059, 0.753171669319272, 0.515425129327923, 0.296175548806787, 0.665178376249969, 0.0218433628324419, 0.0643679571803659, 0.701221072580665, 0.488211358897388, 0.8214950656984, 0.811346856644377, 0.83618796360679, 0.0748408066574484, 0.436013287631795, 0.129921660060063, 0.684180218027905, 0.551822539651766, 0.233757417649031, 0.243846385506913, 0.275729882763699, 0.506465035490692, 0.71488185483031, 0.651927744038403, 0.242717564338818, 0.14213726692833, 0.0819436646997929, 0.256358972750604, 0.0963108928408474, 0.961240731645375, 0.809286808827892, 0.413279631407931, 0.901707275537774, 0.707660013344139, 0.105974593898281, 0.348627322353423, 0.637093344004825, 0.19672862556763, 0.470974079798907, 0.9470642847009, 0.164777805330232, 0.0123005122877657, 0.417783401207998, 0.701012342469767, 0.628121942747384, 0.798788364045322, 0.779734569601715, 0.48669950524345, 0.788949345005676, 0.0846157157793641, 0.830748803447932, 0.0607781691942364, 0.130826523061842, 0.222759227501228, 0.174673525616527, 0.279086240800098, 0.00956874736584723, 0.85863740905188, 0.788739495445043, 0.0806047464720905, 0.74420299869962, 0.363752735313028, 0.346387385157868, 0.710118147544563, 0.181700645480305, 0.117440408095717, 0.65426440956071, 0.481578467646614, 0.491025239229202, 0.288673118222505, 0.791743596550077, 0.362587624695152, 0.903078499715775, 0.112143587553874, 0.597079820232466, 0.191130767809227, 0.00535943638533354, 0.527724357089028, 0.214786684839055, 0.151305745821446, 0.639824600191787, 0.308729373617098, 0.58312981785275, 0.936885085422546, 0.945829240838066, 0.541444980306551, 0.0297637921757996, 0.37268123915419, 0.710365098202601, 0.691670642001554, 0.408524474594742, 0.825607906328514, 0.339663837105036, 0.842816317919642, 0.868788451189175, 0.501800248632208, 0.117198562016711, 0.92786450847052, 0.747157013509423, 0.992315405281261, 0.982038003159687, 0.118019565474242, 0.0609488789923489, 0.198931826045737, 0.275817700894549, 0.0519942992832512, 0.909608739893883, 0.364263652823865, 0.55558655760251, 0.237174928653985, 0.63416043901816, 0.0568105720449239, 0.851477069780231, 0.384305161656812, 0.463852752931416, 0.858298321487382, 0.0473042938392609, 0.0269008059985936, 0.345847638556734, 0.920803658431396, 0.290129241999239, 0.297519785817713, 0.128426727140322, 0.345442177960649, 0.851831204723567, 0.239299239823595, 0.2381048114039, 0.0478104376234114, 0.491026196163148, 0.855440476210788, 0.261030024150386, 0.296676855534315, 0.210572981741279, 0.735584112582728, 0.772885624086484, 0.24800839764066, 0.0646458156406879, 0.771707377163693, 0.0463490455877036, 0.913677687989548, 0.373899368103594, 0.241746897576377, 0.723927803803235, 0.783181084087119, 0.175849460996687, 0.632174374070019, 0.80084946192801, 0.746774468105286, 0.578357151709497, 0.592616725945845, 0.576512907398865, 0.624121158616617, 0.252701843623072, 0.108312303666025, 0.6223298269324, 0.822475476423278, 0.0710640852339566, 0.318196430802345, 0.0639208068605512, 0.275776538532227, 0.972694515949115, 0.246643337886781, 0.606074046576396, 0.901570956222713, 0.780277550918981, 0.780631880275905, 0.999298150651157, 0.689135240158066, 0.443682828685269, 0.104566480731592, 0.527330462122336, 0.843789347913116, 0.81550676561892, 0.50546301738359, 0.142526938347146, 0.912561961682513, 0.585637249052525, 0.982873171102256, 0.945662259124219, 0.0770867180544883, 0.619147780584171, 0.172921100398526, 0.490967937279493, 0.108842365443707, 0.079292330192402, 0.440362084889784, 0.87042427668348, 0.942651101620868, 0.508829540573061, 0.892071280395612, 0.989000204252079, 0.980715289711952, 0.606098177144304, 0.068806363735348, 0.928670949768275, 0.361153508070856, 0.341791198123246, 0.963479431578889, 0.904001348186284, 0.661521149799228, 0.922575515229255, 0.434324811212718, 0.00317672872915864, 0.466554336482659, 0.198967120610178, 0.148682400817052, 0.111608375096694, 0.32805879577063, 0.862434265669435, 0.465708905598149, 0.704408125719056, 0.41668174858205, 0.590017180191353, 0.279430256923661, 0.618970901938155, 0.168379888636991, 0.434617090970278, 0.570122791919857, 0.347458716249093, 0.0238481680862606, 0.840786589076743, 0.0551841240376234, 0.757552344817668, 0.963577720569447, 0.721702698152512, 0.312142421491444, 0.0844512083567679, 0.474287107121199, 0.544129100861028, 0.311754728201777, 0.47877496574074, 0.0551223882939667, 0.36832904140465, 0.776001127203926, 0.161192260915413, 0.625964696519077, 0.773267855169252, 0.225879716221243, 0.968611396150663, 0.57617946062237, 0.0418768425006419, 0.0193576882593334, 0.141839291900396, 0.3369405050762, 0.316256830468774, 0.0589736697729677, 0.502794692991301, 0.546186763560399, 0.784798054024577, 0.716855519916862, 0.872375045670196, 0.301780732814223, 0.210972340544686, 0.736033690162003, 0.859148915158585, 0.89553645462729, 0.277956271544099, 0.577418127562851, 0.785983917769045, 0.70479243225418, 0.283708619652316, 0.564200719352812, 0.0049965875223279, 0.340174331329763, 0.0252893278375268, 0.808723308378831, 0.459276563720778, 0.681324324337766, 0.31290883012116, 0.0934819439426064, 0.82056212821044, 0.0813766303472221, 0.340162563603371, 0.268692648271099, 0.981373681919649, 0.345527132973075, 0.533007003366947, 0.477060297969729, 0.924996972549707, 0.944847214501351, 0.567561614327133, 0.289300374686718, 0.53640012582764, 0.603489762404934, 0.643917164998129, 0.832187559222803, 0.665012933546677, 0.279098288156092, 0.441330534406006, 0.762261861469597, 0.288698439253494, 0.615388927282766, 0.82071556802839, 0.787007842212915, 0.914949729572982, 0.171941422158852), actual = c(1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.763546127825975, 0.5892770125065, 0.321430875454098, 0.197732938453555, 0.508014852646738, 0.465398274594918, 0.810686303302646, 0.790019976207986, 0.88007167330943, 0.966706288978457, 0.886661121621728, 0.322934268042445, 0.661922901170328, 0.0341774348635226, 0.256075812503695, 0.130825235508382, 0.219143501948565, 0.619111957959831, 0.45391710428521, 0.530573654454201, 0.128656739369035, 0.450055372202769, 0.628751517739147, 0.0807524519041181, 0.0571259437128901, 0.0755339644383639, 0.945118042873219, 0.565865303156897, 0.8379347959999, 0.823223248822615, 0.42270437348634, 0.823500659782439, 0.270751625066623, 0.526329410960898, 0.305273233214393, 0.86198786762543, 0.979644102975726, 0.718538439366966, 0.511576334480196, 0.486231773160398, 0.761738611618057, 0.99425572482869, 0.139545602723956, 0.712091342778876, 0.369718969101086, 0.496238953899592, 0.659944700542837, 0.475570633541793, 0.113366656238213, 0.555719781434163, 0.148505948483944, 0.592308616964146, 0.0169910178519785, 0.401842115446925, 0.55323847848922, 0.632384925149381, 0.909023829037324, 0.762676511192694, 0.0115249534137547, 0.779044610681012, 0.0458323284983635, 0.900079345330596, 0.38134220871143, 0.187556519405916, 0.659005098510534, 0.605101145338267, 0.291366076329723, 0.904078568564728, 0.588596215471625, 0.525060548447073, 0.560825073858723, 0.159014951437712, 0.677372099366039, 0.309948878595605, 0.724589133169502, 0.226636696839705, 0.639094712911174, 0.931175634963438, 0.61778705380857, 0.0888845240697265, 0.328035628655925, 0.402021084912121, 0.837739593582228, 0.90360128111206, 0.423284430988133, 0.645881532225758, 0.00574111705645919, 0.462283306987956, 0.931833227630705, 0.36426384979859, 0.957534728804603, 0.95060638198629, 0.380082550924271, 0.35023205075413, 0.521890452131629, 0.525783349759877, 0.367533826734871, 0.435810886090621, 0.763826416805387, 0.738072886830196, 0.836291002575308, 0.0406404540408403, 0.219663369003683, 0.414480363717303, 0.815672558033839, 0.6660734442994, 0.0907287204172462, 0.69847202533856, 0.844414460239932, 0.498586900299415, 0.391513736918569, 0.633852813858539, 0.599197877803817, 0.889442907879129, 0.146488559432328, 0.126126856775954, 0.371944836573675, 0.569316594861448, 0.687858405988663, 0.966862039174885, 0.480744641274214, 0.0078874547034502, 0.811201547272503, 0.13977694651112, 0.884456095751375, 0.876017505535856, 0.0607608372811228, 0.783963415073231, 0.42401241324842, 0.290977294556797, 0.554563872050494, 0.293990285368636, 0.470058660255745, 0.742273327196017, 0.229399919742718, 0.340913910651579, 0.331299512647092, 0.355092127574608, 0.139790613204241, 0.562674363842234, 0.768226486863568, 0.609559469390661, 0.220672159222886, 0.182671225164086, 0.272433748003095, 0.466586308088154, 0.581955836387351, 0.518663407070562, 0.850500227883458, 0.297089341795072, 0.0328742146957666, 0.00309094646945596, 0.926568799652159, 0.0268947000149637, 0.831422966206446, 0.0577563918195665, 0.563744208076969, 0.536627464229241, 0.150170453824103, 0.801373488502577, 0.749572403263301, 0.266438841586933, 0.156650827731937, 0.280621591256931, 0.00867412565276027, 0.229779471177608, 0.56750028510578, 0.614564846502617, 0.172900061355904, 0.176138321869075, 0.994255888508633, 0.383954712422565, 0.786159516777843, 0.350472702644765, 0.806094910949469, 0.609775190008804, 0.677916075335816, 0.950382420327514, 0.578926568618044, 0.396057379897684, 0.875269905896857, 0.952011796878651, 0.00669696251861751, 0.720803852891549, 0.577018854906783, 0.09147545741871, 0.765443310840055, 0.446322152856737, 0.166705233044922, 0.767604083986953, 0.695611235918477, 0.312139007495716, 0.631416189018637, 0.454479895764962, 0.675860702292994, 0.495574481086805, 0.18420501309447, 0.83960204012692, 0.00939364498481154, 0.752045156667009, 0.29583610012196, 0.632433439139277, 0.0164199839346111, 0.14563299738802, 0.439106266479939, 0.907756656641141, 0.435935897519812, 0.979322092374787, 0.983329490059987, 0.895031863823533, 0.894404124701396, 0.91287366556935, 0.981549510266632, 0.418570214882493, 0.772575963288546, 0.974759724456817, 0.0620782969053835, 0.386630217079073, 0.0193584971129894, 0.994597604963928, 0.397234778152779, 0.536947699496523, 0.756670008879155, 0.853637653170153, 0.580848709214479, 0.174763039452955, 0.561911639757454, 0.571956653380767, 0.650336153339595, 0.00956978811882436, 0.599176965421066, 0.496928938664496, 0.0777348186820745, 0.130886554718018, 0.0804176682140678, 0.995755642652512, 0.987224643584341, 0.862237278139219, 0.606334333773702, 0.675304080126807, 0.165554944658652, 0.995024991920218, 0.254532192833722, 0.996247876668349, 0.0187865816988051, 0.544777371920645, 0.375148584134877, 0.214815570507199, 0.66238064179197, 0.44029181660153, 0.403162474744022, 0.535098091233522, 0.146385651547462, 0.378609481267631, 0.924699042225257, 0.668206016765907, 0.524467958137393, 0.35014572320506, 0.682577526429668, 0.551039488753304, 0.308715333696455, 0.664877871517092, 0.78681448684074, 0.792934185126796, 0.578117226017639, 0.940192302688956, 0.1750503545627, 0.397869586478919, 0.361995758255944, 0.71106797689572, 0.508130109868944, 0.169582272646949, 0.483318010577932, 0.627039853483438, 0.901505370857194, 0.605278852395713, 0.55960572254844, 0.652324557304382, 0.00411240034736693, 0.653256789082661, 0.801382505102083, 0.854076700517908, 0.954080459196121, 0.485036414349452, 0.0128645531367511, 0.354530378710479, 0.735510215163231, 0.323976139770821, 0.225879246601835, 0.309006666298956, 0.496096086222678, 0.644921586848795, 0.222880312707275, 0.817721229745075, 0.3663557136897, 0.20093565643765, 0.625844852300361, 0.609297158196568, 0.932158626848832, 0.443174622254446, 0.107227632310241, 0.31264738086611, 0.984278457472101, 0.714874366531149, 0.869471303187311, 0.716794124105945, 0.0369580034166574, 0.162146302405745, 0.213582505006343, 0.0088983706664294, 0.933304209494963, 0.309308216441423, 0.324184520170093, 0.609236736316234, 0.577349988743663, 0.10057469829917, 0.151951237116009, 0.120713697047904, 0.0646503816824406, 0.00129852909594774, 0.0646341145038605, 0.37016060622409, 0.346852717455477, 0.524953369982541, 0.894339418970048, 0.106198891066015, 0.957727104658261, 0.393926287069917, 0.301222959067672, 0.935528106288984, 0.609086781507358, 0.364180359756574, 0.632727917050943, 0.466279809828848, 0.531558316899464, 0.98067480023019, 0.0255275995004922, 0.0880915974266827, 0.134563216939569, 0.667410460300744, 0.218905367190018, 0.70585295627825, 0.77998528117314, 0.764150052797049, 0.465038685360923, 0.584782543359324, 0.691889741225168, 0.298122031847015, 0.837700417498127, 0.93557481886819, 0.661630075192079, 0.555055518401787, 0.378087416756898, 0.958759502507746, 0.55260488204658, 0.324052918236703, 0.989227392477915, 0.632366636767983, 0.888821620028466, 0.0884000337682664, 0.556209929753095, 0.526068036444485, 0.4243502211757, 0.655455479864031, 0.37076539453119, 0.619551487034187, 0.541074552340433, 0.309383958345279, 0.258523928700015, 0.204540719976649, 0.369121063500643, 0.270717054838315, 0.437446608906612, 0.924293785123155, 0.190955689642578, 0.800493650604039, 0.577969639794901, 0.942611432168633, 0.959804177749902, 0.519972363719717, 0.630454314639792, 0.93880985560827, 0.204798671416938, 0.887756110401824, 0.847093979362398, 0.901092590298504, 0.450001569464803, 0.137940211454406, 0.660325936274603, 0.465440609259531, 0.971511917188764, 0.684393763309345, 0.257436173269525, 0.97830322291702, 0.640359236160293, 0.128491975134239, 0.15596982720308, 0.166442565154284, 0.229928129119799, 0.178858715808019, 0.62105548591353, 0.645422692876309, 0.466402303427458, 0.370574701344594, 0.475403263466433, 0.678734426153824, 0.176193571416661, 0.13526131096296, 0.889396321959794, 0.0486900145187974, 0.791365451877937, 0.536700097844005, 0.860509922495112, 0.315729459980503, 0.403820700244978, 0.776947301346809, 0.954735989915207, 0.281666807830334, 0.406408386770636, 0.43568349792622, 0.557737144641578, 0.212172460975125, 0.621525649679825, 0.996563628548756, 0.0865103167016059, 0.558230901136994, 0.651597868883982, 0.568257030565292, 0.160572695080191, 0.0623224573209882, 0.626685288967565, 0.667399110505357, 0.835207245778292, 0.0553945291321725, 0.50085772597231, 0.383293382590637, 0.604463706957176, 0.537967027397826, 0.586267759092152, 0.31713610724546, 0.00309375789947808, 0.855234239948913, 0.802758687641472, 0.715066224336624, 0.988198620965704, 0.776003661798313, 0.880176614038646, 0.617562669562176, 0.356634697411209, 0.172366766957566, 0.315661916974932, 0.857090497622266, 0.854554623598233, 0.363573526032269, 0.955705389147624, 0.47729796054773, 0.351611847057939, 0.159266334958375, 0.661462030140683, 0.605007958365604, 0.205637698760256, 0.153862539213151, 0.267780829453841, 0.746987550985068, 0.748549897223711, 0.599580427398905, 0.348684407770634, 0.312908807769418, 0.0267840749584138, 0.917362653184682, 0.786652852548286, 0.74370218697004, 0.424105393467471, 0.0666361965704709, 0.849958934821188, 0.317946152761579, 0.335925260325894, 0.624453553697094, 0.315070792334154, 0.719857672229409, 0.918460167711601, 0.412532409653068, 0.526555170305073, 0.137976203812286, 0.189491258468479, 0.552338656969368, 0.652315027080476, 0.931182486005127, 0.825727020623162, 0.704955260735005, 0.262669394724071, 0.517330157803372, 0.589729228056967, 0.568403539946303, 0.351113317301497, 0.259084571385756, 0.0683998197782785, 0.382097505033016, 0.513014713767916, 0.137644859729335, 0.997776177246124, 0.64892779616639, 0.424745260970667, 0.823101204819977, 0.226386768743396, 0.368098591221496, 0.443853017408401, 0.979119476862252, 0.497766991378739, 0.956423592055216, 0.022529597626999, 0.175894107436761, 0.834598688874394, 0.552948252763599, 0.19716861192137, 0.135820209281519, 0.427790325833485, 0.559924405068159, 0.507694164523855, 0.737658682279289, 0.468819134635851, 0.495944243157282, 0.0863026019651443, 0.126765455119312, 0.888297348050401, 0.5433768206276, 0.84032569383271, 0.167998296907172, 0.862750827567652, 0.529729147674516, 0.977789528435096, 0.357984951930121, 0.424654462607577, 0.70525075099431, 0.259402308613062, 0.601153739728034, 0.263077588286251, 0.975719164358452, 0.610720267519355, 0.117305330000818, 0.674428093014285, 0.773249749559909, 0.803950067842379, 0.805992650566623, 0.83913164306432, 0.953988405875862, 0.106398070696741, 0.531504172598943, 0.208010682603344, 0.433443099958822, 0.54633837309666, 0.349715634714812, 0.375786405522376, 0.392575346399099, 0.966726414160803, 0.941742558730766, 0.105418711202219, 0.19851855118759, 0.161566011141986, 0.501468709902838, 0.0800512267742306, 0.955502746393904, 0.948387656826526, 0.877090188441798, 0.70918904710561, 0.650381233077496, 0.339550466742367, 0.0511188083328307, 0.29171902127564, 0.514090757118538, 0.411705747013912, 0.366269613150507, 0.252225684700534, 0.669248472200707, 0.817583940690383, 0.478765635984018, 0.843861317262053, 0.327984733274207, 0.752717768307775, 0.517795717343688, 0.935071556828916, 0.21535000228323, 0.970601728884503, 0.4312443069648, 0.275647032773122, 0.254203041084111, 0.750139772891998, 0.318973236950114, 0.878949205158278, 0.752533032791689, 0.879052731208503, 0.0413567763753235, 0.355428769486025, 0.412717374972999, 0.331467237090692, 0.33520764997229, 0.958055147202685, 0.558467507362366, 0.0796934021636844, 0.931763202184811, 0.181316139409319, 0.971272906986997, 0.228193083778024, 0.786206867545843, 0.000332209747284651, 0.690235145622864, 0.987620477098972, 0.306440972024575, 0.667273848783225, 0.139823760604486, 0.724010758567601, 0.215637137182057, 0.26675624283962, 0.301021998981014, 0.110283155692741, 0.871474652318284, 0.184165418380871, 0.525896261679009, 0.518275226233527, 0.786297429585829, 0.942953134188429, 0.802006711484864, 0.428514695959166, 0.13416381762363, 0.84195411670953, 0.606653875205666, 0.827607515268028, 0.69305669143796, 0.3652701722458, 0.442524678772315, 0.936372160678729, 0.42435922450386, 0.0355239582713693, 0.764033226761967, 0.949249176774174, 0.399094629101455, 0.145641650771722, 0.802849275525659, 0.293675769586116, 0.577593898167834, 0.332073011435568, 0.0605656811967492, 0.194972242228687, 0.518799657467753, 0.968981123296544, 0.945626322180033, 0.605899554677308, 0.870203252881765, 0.947895793477073, 0.579401878640056, 0.399897351628169, 0.234144518151879, 0.373681725701317, 0.212530051358044, 0.391969939926639, 0.383055007318035, 0.789566912688315, 0.940933402162045, 0.285087752854452, 0.417097933124751, 0.414368794765323, 0.660820575198159, 0.940745507599786, 0.389025424141437, 0.68058678181842, 0.125151437008753, 0.321663693990558, 0.363400625763461, 0.857102936832234, 0.181888792663813, 0.503452828619629, 0.00930868112482131, 0.993966360809281, 0.841448668856174, 0.831030733184889, 0.314354948932305, 0.271950897527859, 0.400626241462305, 0.201985304476693, 0.0806825531180948, 0.431415849132463, 0.699457755545154, 0.686666994122788, 0.420443970244378, 0.732537071220577, 0.0314536362420768, 0.782579943537712, 0.456054670270532, 0.587596824159846, 0.0994167439639568, 0.302923667011783, 0.985013442812487, 0.356318108504638, 0.71862052497454, 0.302237970521674, 0.939209830481559, 0.106125865364447, 0.263326349901035, 0.24360806401819, 0.598679885268211, 0.659519228851423, 0.125046978006139, 0.0765821952372789, 0.667406833963469, 0.718014209764078, 0.428693261230364, 0.781761375954375, 0.196385422488675, 0.406718875980005, 0.573577303206548, 0.834959691390395, 0.831070537678897, 0.328335667494684, 0.736117670079693, 0.599480828037485, 0.391675173072144, 0.120614877901971, 0.988957522204146, 0.0813424927182496, 0.0398456843104213, 0.629857019754127, 0.24962752382271, 0.214387639425695, 0.0677848374471068, 0.798675034893677, 0.552118510240689, 0.0379507171455771, 0.474067298229784, 0.509752558078617, 0.342303782468662, 0.275382375810295, 0.855409179581329, 0.220742780482396, 0.280577861703932, 0.604826489696279, 0.882278489181772, 0.924592850962654, 0.572341791819781, 0.460587208624929, 0.560790567426011, 0.889392168959603, 0.133510439889506, 0.671567822806537, 0.402624451555312, 0.216782480012625, 0.733768917154521, 0.230928272940218, 0.0687601384706795, 0.99909314699471, 0.844067174009979, 0.180588271003217, 0.206307195592672, 0.0898098654579371, 0.486991977319121, 0.797082133125514, 0.661719681229442, 0.108317815698683, 0.327276328345761, 0.608393496600911, 0.612732208101079, 0.152720435988158, 0.118914994876832, 0.907836656551808, 0.107008773600683, 0.503523220773786, 0.475710378726944, 0.766019047936425, 0.581485795089975, 0.789601766038686, 0.395817334298044, 0.763390709878877, 0.12825554003939, 0.535691322060302, 0.57573676831089, 0.559606818016618, 0.761083803838119, 0.955457379342988, 0.481666060863063, 0.176979163428769, 0.776238634251058, 0.456939540570602, 0.987336614634842, 0.875734966248274, 0.225771174300462, 0.387623032322153, 0.449148317566141, 0.746323792729527, 0.185337180737406, 0.500152222812176, 0.222584744915366, 0.973342768382281, 0.564791427226737, 0.178977696225047, 0.925875263754278, 0.204863086575642, 0.0659822504967451, 0.887378582730889, 0.898940197424963, 0.688866184558719, 0.62174595403485, 0.151529308641329, 0.541610710555688, 0.812239474616945, 0.629846674855798, 0.545612041139975, 0.0959525762591511, 0.564631981542334, 0.0482096513733268, 0.219273532973602, 0.87234260304831, 0.971608402440324, 0.236590087413788, 0.94278856087476, 0.245097478153184, 0.969132390338928, 0.330575932282954, 0.964589734561741, 0.518377437489107, 0.0536003736779094, 0.356466491241008, 0.72584147262387, 0.188165084458888, 0.0256194276735187, 0.416427894961089, 0.0384559039957821, 0.970904325600713, 0.741908272728324, 0.476180909434333, 0.0842246310785413, 0.778166792821139, 0.0343240310903639, 0.0946221691556275, 0.565377159742638, 0.135156202595681, 0.3198127374053, 0.0298799630254507, 0.958115448476747, 0.0750919827260077, 0.842822127044201, 0.632499005645514, 0.648405860876665, 0.340984435984865, 0.372822619741783, 0.546353585785255, 0.973616309463978, 0.0848526074551046, 0.796071351505816, 0.112629902083427, 0.220857254229486, 0.189697779016569, 0.00846389890648425, 0.301206497708336, 0.853793311864138, 0.578177746618167, 0.736192889977247, 0.751540568657219, 0.248926614178345, 0.228618361754343, 0.850099295843393, 0.817940314533189, 0.60909393359907, 0.892753717489541, 0.297472366597503, 0.999992612516508, 0.548258037306368, 0.681034485111013, 0.617611908819526, 0.255229556001723, 0.994234634796157, 0.918696850305423, 0.0665929687675089, 0.133414113661274, 0.807813302613795, 0.643505278043449, 0.0261021601036191, 0.088277714792639, 0.583382211858407, 0.191344574093819, 0.181978867854923, 0.783414323581383, 0.0799684270750731, 0.499238784890622, 0.450225987005979, 0.0500524998642504, 0.906816170783713, 0.353640376357362, 0.193187916651368, 0.532972330460325, 0.0897679626941681, 0.90436685574241, 0.997796453768387, 0.927031082799658, 0.353673405013978, 0.460587363457307, 0.742689688457176, 0.120370797580108, 0.396258114138618, 0.935306803323328, 0.719147581141442, 0.945116720860824, 0.118237439310178, 0.452994076302275, 0.568859806517139, 0.218218622263521, 0.0635181644465774, 0.106186392949894, 0.0518563971854746, 0.195057024480775, 0.465088311582804, 0.588342818897218, 0.794245299650356, 0.579540974693373, 0.353775245370343, 0.127581716980785, 0.872702301014215, 0.0694181635044515, 0.655104621080682, 0.109779756283388, 0.289936474524438, 0.100403926102445, 0.143798526376486, 0.539760749554262, 0.976501783356071, 0.185527859488502, 0.203959526726976, 0.740160466870293, 0.266968545038253, 0.549358125310391, 0.466064362088218, 0.768508410779759, 0.903894591610879, 0.443838571431115, 0.0950166105758399, 0.401284928666428, 0.851177201140672, 0.625015447614715, 0.442878334783018, 0.510481093078852, 0.571350711164996, 0.0983170967083424, 0.590263339225203, 0.09601227985695, 0.0166634570341557, 0.583732362370938, 0.582405675668269, 0.398900107247755, 0.285984749207273, 0.765378333861008, 0.0569920996204019, 0.26836305949837, 0.456861544633284, 0.796567843295634, 0.565387625945732, 0.714437827002257, 0.284371255664155, 0.754701501689851, 0.864042263943702, 0.686031129211187, 0.0695064102765173, 0.918125924421474, 0.0334367610048503, 0.521264559123665, 0.555920260725543, 0.407579679274932, 0.236717195948586, 0.0866354459431022, 0.15228874492459, 0.224228108301759, 0.250905752647668, 0.189004254760221, 0.543846277520061, 0.833531160373241, 0.0622714022174478, 0.175435767043382, 0.121686875354499, 0.750849085859954, 0.166783267864957, 0.449980169069022, 0.167822693008929, 0.999418004415929, 0.895682409638539, 0.69437129679136, 0.733465285971761, 0.629990072455257, 0.573783781845123, 0.327137049054727, 0.21383249363862, 0.683979460038245, 0.926600816193968, 0.448303388198838, 0.522415051469579, 0.960025457432494, 0.513604715932161, 0.241570268524811, 0.324933431111276, 0.934825336327776, 0.0552300103008747, 0.796822667121887, 0.336308281868696, 0.999413876561448), actual = c(0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"