Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.0397541        0.0414891 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5411838        0.5411838        0.4933569 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.405805718153715, 0.970531848724931, 0.15442160028033, 0.702491048257798, 0.894902059342712, 0.0959317192900926, 0.840042944764718, 0.858637457247823, 0.291910845087841, 0.271084189182147, 0.045903675025329, 0.924564264481887, 0.182007444323972, 0.384329596301541, 0.269669079920277, 0.00975769036449492, 0.174058370990679, 0.240379127208143, 0.707149129826576, 0.324640315724537, 0.238541111815721, 0.11722720018588, 0.268598357215524, 0.438352190656587, 0.763487725518644, 0.617480813991278, 0.0470717425923795, 0.215749990660697, 0.349205241072923, 0.544756316347048, 0.454041708959267, 0.653220624662936, 0.448960426263511, 0.393962227972224, 0.981097828131169, 0.282635852927342, 0.132581321056932, 0.314739051274955, 0.681250494206324, 0.114342858083546, 0.194558621384203, 0.379303408320993, 0.896663118153811, 0.0777078557293862, 0.54357085749507, 0.828692110721022, 0.749921462731436, 0.999824891565368, 0.526776376646012, 0.881447384366766, 0.51481819874607, 0.845485791098326, 0.806412927107885, 0.245184456696734, 0.703294758219272, 0.810849752277136, 0.447706853970885, 0.366486134938896, 0.437373536871746, 0.864560666959733, 0.777478683739901, 0.823634282220155, 0.232356215594336, 0.869043708546087, 0.00986314914189279, 0.758106153225526, 0.801179552450776, 0.127870928728953, 0.326638789149001, 0.967743381392211, 0.846426020143554, 0.392177890986204, 0.387214089510962, 0.968337933532894, 0.361788238398731, 0.0229712061118335, 0.81472629844211, 0.586850383551791, 0.158621688606218, 0.0885760912206024, 0.043728994904086, 0.0860616818536073, 0.664480957435444, 0.129107011249289, 0.491435716859996, 0.0442247355822474, 0.198907762533054, 0.824001967906952, 0.649897149531171, 0.0409347151871771, 0.0977610680274665, 0.883510329760611, 0.786909701302648, 0.0350363212637603, 0.211888049030676, 0.564064621459693, 0.650937147205696, 0.0744328834116459, 0.99439162760973, 0.256505002966151, 0.117605139734223, 0.159454548731446, 0.135411735856906, 0.488672646926716, 0.524450517492369, 0.494956238428131, 0.54417353705503, 0.221509063616395, 0.463062347611412, 0.210463980678469, 0.975507852854207, 0.843130674911663, 0.960976760601625, 0.230437230784446, 0.567353701218963, 0.628134530270472, 0.86766107310541, 0.807870670454577, 0.66361508728005, 0.218945140950382, 0.627721536904573, 0.172415043460205, 0.149090345483273, 0.847154163988307, 0.880323011428118, 0.516380169428885, 0.914598799543455, 0.264985139481723, 0.512017039116472, 0.783401993336156, 0.348408099263906, 0.361087720608339, 0.816407069563866, 0.609418930951506, 0.928951482055709, 0.57516234787181, 0.805733320768923, 0.161357996985316, 0.519194537773728, 0.129470437532291, 0.731427199440077, 0.706752124708146, 0.431747477734461, 0.534942982718349, 0.0253286564256996, 0.596496574347839, 0.102586527820677, 0.879814974265173, 0.120151033857837, 0.58887575333938, 0.425960330758244, 0.119219066342339, 0.188796718837693, 0.985416713869199, 0.312289314810187, 0.743399558356032, 0.228249308886006, 0.627844675676897, 0.828236360568553, 0.945637741824612, 0.465398136992007, 0.513965365011245, 0.699651850853115, 0.541404943680391, 0.514540761709213, 0.1458097146824, 0.202711934456602, 0.156677272403613, 0.317587364697829, 0.647492955904454, 0.175267733866349, 0.0723222882952541, 0.646084613166749, 0.67317205783911, 0.108458243077621, 0.877189207589254, 0.242294927127659, 0.981977037154138, 0.741198114585131, 0.444332605693489, 0.399597900453955, 0.0553334376309067, 0.690056536113843, 0.501579596428201, 0.555870796088129, 0.400833549443632, 0.458095423877239, 0.696628172881901, 0.98915262054652, 0.790055701741949, 0.202262127771974, 0.334810232976452, 0.23637406504713, 0.531385489040986, 0.789738417370245, 0.583064464386553, 0.343391096452251, 0.829560111975297, 0.652215645648539, 0.992816030047834, 0.830236875684932, 0.214678288903087, 0.616029945435002, 0.735475622816011, 0.228158100740984, 0.785476814256981, 0.968981103505939, 0.979858120903373, 0.305008763214573, 0.572122578043491, 0.0528588171582669, 0.910783232888207, 0.601643427507952, 0.343343766173348, 0.535617148270831, 0.919998250203207, 0.890066795283929, 0.189424399985, 0.850831177551299, 0.21274346136488, 0.861383436247706, 0.792942468076944, 0.40260594454594, 0.568771089194342, 0.594312857370824, 0.0284880506806076, 0.844737007748336, 0.711558806477115, 0.721280531957746, 0.836512624053285, 0.232640908099711, 0.937198952073231, 0.643697053892538, 0.660757664125413, 0.272203035186976, 0.100694684777409, 0.97285188594833, 0.84194158250466, 0.544334572274238, 0.649157100124285, 0.87617271556519, 0.0432167744729668, 0.658736005658284, 0.0118742457125336, 0.936976304277778, 0.502313239499927, 0.50874076411128, 0.0385494746733457, 0.265209155390039, 0.127169623272493, 0.465811133151874, 0.129022945184261, 0.997194116469473, 0.659313358133659, 0.899889597203583, 0.374619525158778, 0.460812114179134, 0.0867083934135735, 0.624557317700237, 0.349586726631969, 0.172151126200333, 0.552431190153584, 0.690834863111377, 0.677983689820394, 0.0870404937304556, 0.643211464164779, 0.917391657363623, 0.92096844734624, 0.59392160596326, 0.0168196242302656, 0.79817978083156, 0.0910382268484682, 0.582396316342056, 0.727712591644377, 0.922926297178492, 0.591534779407084, 0.173667876049876, 0.626645694021136, 0.589897707104683, 0.829385013552383, 0.673270486528054, 0.408160453196615, 0.901103334967047, 0.319561943178996, 0.00461460626684129, 0.535215413896367, 0.127149810316041, 0.819780811434612, 0.469426911789924, 0.806628680322319, 0.668532972922549, 0.0236312586348504, 0.347908664261922, 0.0307450217660517, 0.752347943838686, 0.290083760395646, 0.232943781651556, 0.285777983488515, 0.876952773192897, 0.978792907902971, 0.610788073856384, 0.108144880039617, 0.394655241398141, 0.102326579857618, 0.157635035226122, 0.945133273489773, 0.0399979820940644, 0.382595088100061, 0.320472110062838, 0.252770093735307, 0.633202751167119, 0.960129676386714, 0.937740032561123, 0.999031914398074, 0.573498205281794, 0.449977866606787, 0.94890758395195, 0.0697030301671475, 0.200049645267427, 0.850480191176757, 0.430372880771756, 0.0596088161692023, 0.61810673493892, 0.054884220007807, 0.573195265838876, 0.752997804433107, 0.237646513152868, 0.110472277738154, 0.806296034017578, 0.114501807373017, 0.706913243280724, 0.152920407475904, 0.489463648293167, 0.0710143477190286, 0.869901063386351, 0.0274966035503894, 0.179695431143045, 0.3155458252877, 0.83448855066672, 0.408401516498998, 0.868673054734245, 0.580721231643111, 0.760438663186505, 0.107144878711551, 0.56536889844574, 0.91309384140186, 0.540103694424033, 0.96422414155677, 0.513933100737631, 0.947197531117126, 0.956530610565096, 0.335432196967304, 0.868660347070545, 0.958745323121548, 0.77559121302329, 0.419545787153766, 0.954674951499328, 0.812746418872848, 0.437978032743558, 0.379867535317317, 0.268420308362693, 0.183553585084155, 0.575699276523665, 0.299749726662412, 0.0401242030784488, 0.900853773346171, 0.261989877559245, 0.616138694342226, 0.704925657249987, 0.00593373668380082, 0.402049264637753, 0.247108459705487, 0.339253776706755, 0.0178156662732363, 0.248370438348502, 0.893370678881183, 0.0300445263274014, 0.499386880779639, 0.370440971339121, 0.733409581240267, 0.630794857395813, 0.639355619437993, 0.822855047415942, 0.398044613655657, 0.931882854783908, 0.73651770898141, 0.528811427531764, 0.956862244522199, 0.51107599074021, 0.554654428036883, 0.575497408630326, 0.316780387656763, 0.675211247755215, 0.923015396110713, 0.949376334669068, 0.44277614261955, 0.747003326658159, 0.341437668073922, 0.908214148832485, 0.267799015156925, 0.173632071819156, 0.432789939921349, 0.846111024962738, 0.691624517086893, 0.373412462882698, 0.166734830476344, 0.112045124173164, 0.355772146489471, 0.118702151812613, 0.482419315492734, 0.0595238094683737, 0.160470210481435, 0.955727837281302, 0.190675920573995, 0.444339132634923, 0.470848519122228, 0.234200159320608, 0.179621563060209, 0.742960804142058, 0.850002795225009, 0.679642957402393, 0.0169363128952682, 0.302985872607678, 0.624349218094721, 0.716377036646008, 0.959820070303977, 0.71152529749088, 0.617528801783919, 0.843439645133913, 0.940142279723659, 0.079857122618705, 0.637025052914396, 0.148941719671711, 0.00486706499941647, 0.634142902679741, 0.687443161150441, 0.038199495524168, 0.540748805273324, 0.605627373093739, 0.0148159910459071, 0.482399915810674, 0.23306889901869, 0.563579968642443, 0.0373402994591743, 0.329670117003843, 0.297534639481455, 0.387320953188464, 0.290000744862482, 0.933555533410981, 0.710972122848034, 0.241855068597943, 0.0881132306531072, 0.634276215685531, 0.584547873586416, 0.0582641672808677, 0.269710410153493, 0.505561148514971, 0.00599643238820136, 0.259063602192327, 0.301913418341428, 0.320421667536721, 0.751385540701449, 0.640201624715701, 0.729605240281671, 0.812501158565283, 0.107418205821887, 0.319773412076756, 0.748502434697002, 0.91400037589483, 0.152527091326192, 0.0726735715288669, 0.890926051419228, 0.0643109302036464, 0.405290741939098, 0.546379734296352, 0.201664617750794, 0.130186018999666, 0.107123075751588, 0.546907945536077, 0.430836010724306, 0.0174205226358026, 0.0419885218143463, 0.298637573374435, 0.443934364244342, 0.670662830816582, 0.862483076518402, 0.156029174337164, 0.588186450535432, 0.520415457896888, 0.53083535679616, 0.636066870298237, 0.427023675525561, 0.515066713327542, 0.188685541972518, 0.934929259587079, 0.943353516981006, 0.911850935779512, 0.222585605224594, 0.372687185415998, 0.976169604342431, 0.158983674366027, 0.123813060810789, 0.787309661274776, 0.0592322056181729, 0.0575485429726541, 0.845802531111985, 0.0189716492313892, 0.0903860235121101, 0.963431917363778, 0.21434650872834, 0.716070955619216, 0.569416569778696, 0.356218826025724, 0.178299488965422, 0.441237232415006, 0.468745702179149, 0.634000687161461, 0.960390894906595, 0.0497575143817812, 0.480142392218113, 0.652799640549347, 0.805837797932327, 0.187540040817112, 0.933151851873845, 0.00214711506851017, 0.950358820147812, 0.306694949744269, 0.432652816874906, 0.0162246918771416, 0.247281013056636, 0.457655594218522, 0.00591596192680299, 0.197126102400944, 0.253074463224038, 0.144091720227152, 0.012766313040629, 0.575934609863907, 0.215644218958914, 0.140321862651035, 0.876393356127664, 0.369521588552743, 0.370737557066604, 0.101963655557483, 0.255770412040874, 0.349105405155569, 0.816485135583207, 0.684419502737001, 0.900518325623125, 0.414040335686877, 0.858538449509069, 0.331352391513065, 0.165927555179223, 0.429437417536974, 0.8850401584059, 0.662238265154883, 0.78291177097708, 0.904034636681899, 0.141317120986059, 0.0469538094475865, 0.257897560019046, 0.596142935100943, 0.571369794430211, 0.362240506336093, 0.511739786481485, 0.226371425902471, 0.00559006445109844, 0.847858038032427, 0.0342882678378373, 0.523303171619773, 0.760682294843718, 0.933718883665279, 0.947265617549419, 0.132944722194225, 0.924480282468721, 0.778672616230324, 0.508264843374491, 0.436687348643318, 0.904254933120683, 0.344881194876507, 0.157191649312153, 0.51166416797787, 0.25226560793817, 0.35193750844337, 0.90250778500922, 0.103798870462924, 0.394368833396584, 0.266249559354037, 0.919935553101823, 0.87781265610829, 0.445559374289587, 0.351626051124185, 0.752918988699093, 0.607877189293504, 0.143262204714119, 0.871817915467545, 0.819999316940084, 0.53831042139791, 0.307762269861996, 0.527446187799796, 0.750047440174967, 0.828207287471741, 0.305091454880312, 0.897600683150813, 0.20638806046918, 0.48271450935863, 0.538117316085845, 0.963620576774701, 0.39686098205857, 0.685132276965305, 0.419124789303169, 0.258742928970605, 0.0889158912468702, 0.0301503648515791, 0.696302060037851, 0.753734424943104, 0.24504533642903, 0.192054885905236, 0.917885083705187, 0.804786834167317, 0.605518020922318, 0.199927949579433, 0.190659486223012, 0.708509764866903, 0.269035011064261, 0.815182078396901, 0.378099760971963, 0.0100608370266855, 0.800525475526229, 0.147959398571402, 0.259778839536011, 0.837440670235083, 0.951523778028786, 0.114699150202796, 0.968864872353151, 0.232426714617759, 0.989307367475703, 0.00603273068554699, 0.541675059124827, 0.664002651115879, 0.603018820285797, 0.701707331929356, 0.439695563400164, 0.918679331662133, 0.937608574982733, 0.666600605938584, 0.329258024459705, 0.372853454668075, 0.903580904239789, 0.371183428447694, 0.074245119234547, 0.152095675701275, 0.519610141403973, 0.204976739361882, 0.58227153099142, 0.169175722170621, 0.658130430383608, 0.16262316564098, 0.309753225883469, 0.380790863186121, 0.934429432731122, 0.223788264440373, 0.661797368433326, 0.221683346433565, 0.800128433620557, 0.163851593621075, 0.61758132930845, 0.613117207307369, 0.246072546346113, 0.364462178666145, 0.5623191841878, 0.659148583887145, 0.0568237390834838, 0.856812929268926, 0.0701317489147186, 0.852749492507428, 0.36654582596384, 0.9714198152069, 0.738128735683858, 0.745123679982498, 0.976249242201447, 0.631896247388795, 0.418335938826203, 0.411565152695403, 0.978822581470013, 0.636614670744166, 0.0718265224713832, 0.98407410364598, 0.87968145404011, 0.154317596694455, 0.340667105047032, 0.364282045746222, 0.670818740036339, 0.133772211382166, 0.172288889531046, 0.695599285885692, 0.427950292360038, 0.218186567071825, 0.275670326082036, 0.144541013520211, 0.706143819261342, 0.0377692775800824, 0.217925930162892, 0.0857135355472565, 0.782195577165112, 0.75404875469394, 0.704474458005279, 0.741348090115935, 0.915770990308374, 0.585975599475205, 0.152718820143491, 0.344816406955943, 0.736966014839709, 0.778362897457555, 0.618011576589197, 0.474291190737858, 0.657096916111186, 0.639232548186556, 0.676138056674972, 0.514302789233625, 0.897618582239375, 0.69184307125397, 0.317332206759602, 0.545077611226588, 0.615796403726563, 0.0862582218833268, 0.956956677138805, 0.287951780017465, 0.664352870546281, 0.0170950144529343, 0.704308126820251, 0.16138845635578, 0.5344669574406, 0.412707517389208, 0.891717260237783, 0.580886108800769, 0.653584676329046, 0.684313483070582, 0.531893666135147, 0.0890490885358304, 0.66309591033496, 0.798352477606386, 0.687837511766702, 0.279914961894974, 0.418678139802068, 0.0792062229011208, 0.637099909130484, 0.0245819168630987, 0.221967109246179, 0.298733000643551, 0.377142023062333, 0.599917635088786, 0.862724025035277, 0.921217736089602, 0.483056289609522, 0.428040329134092, 0.0922425405588001, 0.143381804926321, 0.167463308665901, 0.682367295725271, 0.74890577653423, 0.631781199714169, 0.379980482859537, 0.908747245324776, 0.0473804117646068, 0.0152938663959503, 0.791961946990341, 0.00224366481415927, 0.822916590375826, 0.281140038510785, 0.399768650298938, 0.0436348873190582, 0.280027330387384, 0.511938779149204, 0.695671379799023, 0.422551260562614, 0.423758725868538, 0.174623985541984, 0.733879431383684, 0.149827337823808, 0.697973608504981, 0.821176039753482, 0.466027122689411, 0.0845295970793813, 0.676104017766193, 0.49506093445234, 0.391294436529279, 0.323836484691128, 0.633696396835148, 0.964042082196102, 0.315950985765085, 0.640009233728051, 0.422694593435153, 0.72213263483718, 0.970939478371292, 0.44482698617503, 0.837599149905145, 0.541268358007073, 0.844642743002623, 0.43772240052931, 0.260351528180763, 0.403155851876363, 0.930371488211676, 0.867075970862061, 0.532700537005439, 0.377323757158592, 0.270985249895602, 0.91189249092713, 0.636061665136367, 0.103244750062004, 0.590479464270175, 0.66830997983925, 0.891817395575345, 0.58750035869889, 0.642413006396964, 0.866894757840782, 0.962910131085664, 0.972348420647904, 0.461795716779307, 0.251590366940945, 0.420722058508545, 0.574431002140045, 0.730114576872438, 0.996818388346583, 0.272907769540325, 0.762705453205854, 0.311775365378708, 0.748640668112785, 0.503257907927036, 0.166273127077147, 0.942837792448699, 0.80347500462085, 0.930851584300399, 0.479120516451076, 0.176818512380123, 0.465321372263134, 0.132550851907581, 0.361266889842227, 0.90466650063172, 0.344586520921439, 0.689808736555278, 0.251483669271693, 0.292028011288494, 0.457600959343836, 0.554524247767404, 0.312180238310248, 0.183618326904252, 0.38838258292526, 0.698280573822558, 0.974409742746502, 0.176327130291611, 0.839739059796557, 0.722036071354523, 0.151331601664424, 0.945475722197443, 0.34718127688393, 0.557410390581936, 0.958216960774735, 0.992831255076453, 0.654978349572048, 0.714669048087671, 0.473737549036741, 0.580358895706013, 0.526809287723154, 0.900881401496008, 0.875279746949673, 0.11173537443392, 0.158656134270132, 0.336743162944913, 0.702463112771511, 0.659336297307163, 0.421999171376228, 0.116540258517489, 0.692714064847678, 0.761429832316935, 0.201345829060301, 0.150648029753938, 0.607405032496899, 0.558403303613886, 0.675310552120209, 0.930296082282439, 0.989563800161704, 0.253517567645758, 0.368903908180073, 0.293120608199388, 0.396105712978169, 0.338985336944461, 0.934011889854446, 0.854978083865717, 0.119669071631506, 0.307413939852268, 0.335245755966753, 0.238009755034, 0.544483318459243, 0.818475056206807, 0.310050812549889, 0.284226424060762, 0.384656245587394, 0.493214144837111, 0.436859502457082, 0.798394272569567, 0.202790220733732, 0.0325627718120813, 0.530994813889265, 0.741515069268644, 0.911338232224807, 0.935591922141612, 0.086311950115487, 0.802064330084249, 0.891811732202768, 0.606893631629646, 0.727706205798313, 0.612595513695851, 0.13582745147869, 0.111892146989703, 0.055067275185138, 0.155712156556547, 0.402181131299585, 0.859423944959417, 0.422574326163158, 0.677076992811635, 0.907587062334642, 0.231504255440086, 0.505586463958025, 0.817773383343592, 0.458771191071719, 0.451581393368542, 0.21365749114193, 0.281234489521012, 0.0650389196816832, 0.203784472309053, 0.278149032266811, 0.862674089148641, 0.599556118715554, 0.357703087152913, 0.474398267455399, 0.942979212617502, 0.999642042908818, 0.895046351011842, 0.584268855629489, 0.73151583224535, 0.0818659719079733, 0.854765910189599, 0.479605421656743, 0.856651163427159, 0.263683722121641, 0.762922141933814, 0.526678540743887, 0.119844657136127, 0.0663042631931603, 0.652807167731225, 0.686520509887487, 0.159974720794708, 0.67889228509739, 0.765908442204818, 0.506785499164835, 0.456159333931282, 0.582277150591835, 0.839213551254943, 0.991005052812397, 0.178485953947529, 0.683741027489305, 0.301241558743641, 0.639879200374708, 0.574913631659001, 0.907604927197099, 0.891804824583232, 0.866434415103868, 0.306516823358834, 0.0183785033877939, 0.053231023484841, 0.297140856040642, 0.656550322892144, 0.889103959547356, 0.865294776158407, 0.237698103301227, 0.0238797594793141, 0.820853507611901, 0.853336660191417, 0.978713396238163, 0.168501818086952, 0.18035076581873, 0.236670017708093, 0.497451066970825, 0.232984417816624, 0.29608476604335, 0.475124339340255, 0.133067204151303, 0.562567159067839, 0.793744358932599, 0.871388390427455, 0.710074550239369, 0.846936988411471, 0.642502661095932, 0.0367536665871739, 0.150597081054002, 0.102901319973171, 0.064438151428476, 0.0900981808081269, 0.63669649627991, 0.930462109623477), actual = c(1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.634916285518557, 0.998341318685561, 0.290672561852261, 0.0272986989002675, 0.520728396717459, 0.220659021986648, 0.27606047783047, 0.831644846359268, 0.82841079053469, 0.586681548738852, 0.955671704607084, 0.819085027324036, 0.0539546804502606, 0.431543913204223, 0.866799930343404, 0.115020615281537, 0.842356516513973, 0.355807354440913, 0.121819004416466, 0.839352580253035, 0.366842234507203, 0.20173985301517, 0.610756394220516, 0.373257164144889, 0.926013329997659, 0.906760244863108, 0.654777564108372, 0.17457686830312, 0.273625633912161, 0.867641928838566, 0.351468834327534, 0.127199738752097, 0.398664859589189, 0.00689459964632988, 0.336908067110926, 0.663163420045748, 0.453689569607377, 0.501289853360504, 0.424083502031863, 0.575997411739081, 0.70400521508418, 0.96284185256809, 0.972048241412267, 0.154089432675391, 0.960668032988906, 0.42318355338648, 0.440832179971039, 0.47596695763059, 0.893459128914401, 0.531066211638972, 0.414881247794256, 0.685634113382548, 0.335675138048828, 0.356554246507585, 0.0401992613915354, 0.637702054344118, 0.343906983966008, 0.0520898888353258, 0.926547937793657, 0.786102933809161, 0.859327149810269, 0.596070976695046, 0.692841448355466, 0.335357055068016, 0.119344398146495, 0.698555089533329, 0.283680234104395, 0.636049712775275, 0.106381449149922, 0.343059111386538, 0.401765564456582, 0.661149799125269, 0.511435843305662, 0.883383541600779, 0.707387186586857, 0.846143331611529, 0.9740015626885, 0.0253599896095693, 0.983129191678017, 0.74971355474554, 0.954118923051283, 0.584518287098035, 0.316577420802787, 0.261935405200347, 0.991965922527015, 0.113379992777482, 0.77019520313479, 0.776707336772233, 0.129847327247262, 0.332452409900725, 0.835565981687978, 0.435683769639581, 0.404042731970549, 0.565934487152845, 0.165442377328873, 0.830163536127657, 0.64303714921698, 0.941169099183753, 0.340534809976816, 0.139914158731699, 0.119990303646773, 0.0514104850590229, 0.889170957030728, 0.473391401814297, 0.841752033447847, 0.673155118478462, 0.642898110672832, 0.105122171342373, 0.204795961501077, 0.331141091883183, 0.779684846289456, 0.851373770041391, 0.50144256092608, 0.973742579575628, 0.35952944192104, 0.0424425487872213, 0.495194221381098, 0.935874241171405, 0.705166691914201, 0.701046030037105, 0.100323024671525, 0.704391256673262, 0.268975649727508, 0.821022830437869, 0.842457780614495, 0.577099005226046, 0.866184828337282, 0.705826415214688, 0.379438433330506, 0.466010287636891, 0.788243992719799, 0.620603335788473, 0.851274438900873, 0.35859426856041, 0.659208963392302, 0.335988014237955, 0.758361828746274, 0.467459986684844, 0.064088381594047, 0.287187223788351, 0.672368091298267, 0.0328254101332277, 0.0290659365709871, 0.151175470091403, 0.679330838145688, 0.985341486055404, 0.238085797289386, 0.854331075912341, 0.764129630988464, 0.0656484023202211, 0.601904697250575, 0.661179449409246, 0.153392813401297, 0.214621288934723, 0.741999668069184, 0.17462545633316, 0.882532799383625, 0.470893860096112, 0.4368116769474, 0.497189236339182, 0.200215953635052, 0.248116934439167, 0.343777680536732, 0.0572592618409544, 0.867825522320345, 0.876283491030335, 0.564090269384906, 0.0117356879636645, 0.359844130929559, 0.265743711264804, 0.0284654677379876, 0.866524307988584, 0.932832665508613, 0.367688355967402, 0.824476612498984, 0.0957604264840484, 0.646199700655416, 0.250919367885217, 0.392091360408813, 0.963949502445757, 0.88931985385716, 0.342502050567418, 0.788783994968981, 0.913564480375499, 0.707761714467779, 0.47658937680535, 0.336709988536313, 0.15122883557342, 0.789627634221688, 0.420971759362146, 0.217721168883145, 0.36900132521987, 0.850928799947724, 0.534586933674291, 0.725466179661453, 0.0555621129460633, 0.257375620072708, 0.753364266129211, 0.637819595402107, 0.102778136497363, 0.397996924817562, 0.97434123326093, 0.964213979663327, 0.424031004542485, 0.729125153971836, 0.0246935589239001, 0.443015002412722, 0.127039790153503, 0.354837791761383, 0.56170171382837, 0.203753997338936, 0.380663407035172, 0.200740791391581, 0.169199517695233, 0.59073777939193, 0.930642237886786, 0.155646757455543, 0.333199812332168, 0.175815546186641, 0.32885319669731, 0.781810790067539, 0.974096327321604, 0.582571880891919, 0.798414580989629, 0.278160928981379, 0.684311913326383, 0.544351896038279, 0.476985676912591, 0.552180916070938, 0.819762353785336, 0.724046811694279, 0.0363505606073886, 0.643418749095872, 0.0865425304509699, 0.508987960871309, 0.178177562076598, 0.612282052170485, 0.948576715309173, 0.273901839740574, 0.37074555340223, 0.334032630780712, 0.803597069112584, 0.0422025066800416, 0.293120692949742, 0.617252755677328, 0.602970820618793, 0.229307270841673, 0.143200458027422, 0.592864206526428, 0.761579616460949, 0.250649610999972, 0.960353877861053, 0.507594270864502, 0.69187812670134, 0.109387404518202, 0.840230953413993, 0.556754436576739, 0.236928512109444, 0.0752930028829724, 0.749469220871106, 0.331457485212013, 0.402055224170908, 0.0947800425346941, 0.538453676505014, 0.412658974295482, 0.521839340450242, 0.890739673981443, 0.310063516721129, 0.417511355364695, 0.260621945839375, 0.77351705590263, 0.416596163529903, 0.41997791477479, 0.636438195826486, 0.350301816826686, 0.286158858099952, 0.660456456476822, 0.701923770131543, 0.818321889499202, 0.851700270548463, 0.0813270658254623, 0.811517447698861, 0.789330063620582, 0.620831134729087, 0.430158610222861, 0.937762811779976, 0.94650224619545, 0.264703825814649, 0.29764675302431, 0.0746433108579367, 0.867384217213839, 0.179180181352422, 0.191525866510347, 0.260154670104384, 0.398813483770937, 0.0864955293945968, 0.147970315301791, 0.396660644095391, 0.543339573778212, 0.285098799970001, 0.574602802051231, 0.27039331314154, 0.648273737402633, 0.679240429773927, 0.381191262509674, 0.670047373510897, 0.915481189498678, 0.684805601835251, 0.60641384171322, 0.394066556356847, 0.71003861585632, 0.417845588875934, 0.705940977903083, 0.522965848911554, 0.412983336020261, 0.882114908192307, 0.059621412307024, 0.0340456038247794, 0.815044496906921, 0.976697652600706, 0.548425932880491, 0.2689826390706, 0.918945868033916, 0.579622474266216, 0.401272594928741, 0.366294831736013, 0.415324452100322, 0.765786916716024, 0.551385273691267, 0.112922079861164, 0.633090696996078, 0.73694295110181, 0.29321322706528, 0.283367942785844, 0.659327375469729, 0.193114930065349, 0.0632867773529142, 0.415944395354018, 0.881683589890599, 0.252050649374723, 0.481841558590531, 0.930171745829284, 0.853661660570651, 0.208512980490923, 0.890493921469897, 0.232154691126198, 0.545797379920259, 0.994509226642549, 0.591438914882019, 0.503759012091905, 0.296686770394444, 0.100203175097704, 0.847137158969417, 0.803179836133495, 0.0762988654896617, 0.658556642010808, 0.490941682830453, 0.822479341877624, 0.980368672637269, 0.598564743297175, 0.959717742167413, 0.470125989755616, 0.972995782736689, 0.257578634889796, 0.872046758420765, 0.80896833864972, 0.524850394343957, 0.83458332833834, 0.0710675013251603, 0.540313920704648, 0.772554714232683, 0.275376744102687, 0.871580314124003, 0.617035163100809, 0.0225827661342919, 0.804906696081161, 0.327192803379148, 0.00778325041756034, 0.395551999332383, 0.801537175662816, 0.790231944760308, 0.881612520199269, 0.94033844047226, 0.653611809713766, 0.887143595376983, 0.731329263420776, 0.527548242127523, 0.831784220179543, 0.0906688431277871, 0.354124394711107, 0.955704265041277, 0.464688972802833, 0.236313014524058, 0.182567363837734, 0.947581845335662, 0.347783808596432, 0.352293798932806, 0.160523922648281, 0.179522194201127, 0.0886351345106959, 0.070561348926276, 0.720653043361381, 0.71290647611022, 0.160605354700238, 0.701639098813757, 0.123280945234001, 0.968828068114817, 0.535522361751646, 0.77937071188353, 0.673060106579214, 0.135299876099452, 0.273702028905973, 0.261671141954139, 0.222679632017389, 0.246415040688589, 0.873268462950364, 0.901103464188054, 0.642629110952839, 0.570771044120193, 0.564875612268224, 0.934214989189059, 0.455167488660663, 0.693243962014094, 0.073271491099149, 0.177002589451149, 0.581839160295203, 0.147017203969881, 0.795051075983793, 0.362274456303567, 0.393318936694413, 0.557071332354099, 0.901199160376564, 0.962796625448391, 0.388700586743653, 0.711893654894084, 0.211394662037492, 0.300641994690523, 0.534627933753654, 0.866685586282983, 0.0879652660805732, 0.527919809566811, 0.951779702678323, 0.413652976974845, 0.743365748319775, 0.80440412601456, 0.10681997728534, 0.356483403826132, 0.740021227858961, 0.9511493826285, 0.474900288274512, 0.39573353715241, 0.396051392890513, 0.645247478969395, 0.996735745342448, 0.849163651233539, 0.0737626859918237, 0.468393182381988, 0.269194294698536, 0.572663782164454, 0.765316508477554, 0.333998437272385, 0.539627775782719, 0.720957519253716, 0.75709009449929, 0.680122988764197, 0.18515609530732, 0.930335640441626, 0.69640837633051, 0.816850945353508, 0.298012235900387, 0.816410155501217, 0.310865790117532, 0.678880555555224, 0.82743624295108, 0.156166438478976, 0.410123397130519, 0.445345004554838, 0.795770905446261, 0.217474440578371, 0.0450463478919119, 0.0405348667409271, 0.976099006133154, 0.220785001292825, 0.621409127255902, 0.0105431915726513, 0.787072110222653, 0.393300761235878, 0.981955129187554, 0.0480064449366182, 0.963730742223561, 0.452317115617916, 0.181954108411446, 0.982161325402558, 0.421743432292715, 0.0830178507603705, 0.392141738673672, 0.490799304097891, 0.949156890623271, 0.353107690345496, 0.19058735971339, 0.695046426495537, 0.478912470862269, 0.404774520313367, 0.626752684591338, 0.612838016822934, 0.707415904384106, 0.296849627513438, 0.226302293129265, 0.168149208417162, 0.346274415729567, 0.596776837715879, 0.260933811543509, 0.59487235173583, 0.264632290927693, 0.164462933316827, 0.513089482905343, 0.279282299336046, 0.527840593131259, 0.0906899783294648, 0.210806557442993, 0.349655590951443, 0.0287859325762838, 0.753988327458501, 0.499224463710561, 0.281838250579312, 0.830651574069634, 0.751720895757899, 0.0295424016658217, 0.535894925473258, 0.278297647368163, 0.16948754619807, 0.065798208815977, 0.448165024863556, 0.299955948255956, 0.894761914853007, 0.0186828824225813, 0.750215200940147, 0.911959408549592, 0.288115417119116, 0.819379579741508, 0.584597848588601, 0.158564841607586, 0.828108242480084, 0.740351951215416, 0.244704905198887, 0.0696057446766645, 0.172434109961614, 0.0494514442980289, 0.689299792516977, 0.856014278484508, 0.128364755539224, 0.0555600330699235, 0.0869113591033965, 0.841394831193611, 0.680551561526954, 0.136253148782998, 0.119934781920165, 0.393928533885628, 0.483573061879724, 0.640401558252051, 0.802899211877957, 0.556973728584126, 0.712501362897456, 0.05838882480748, 0.51455966453068, 0.402133301598951, 0.0844879420474172, 0.999748817412183, 0.287442690925673, 0.893370825331658, 0.70648033823818, 0.348991652252153, 0.269639106467366, 0.093999347416684, 0.869346303399652, 0.0556240014266223, 0.95545122702606, 0.445448130602017, 0.279517411487177, 0.0697583961300552, 0.0747780846431851, 0.266384235816076, 0.746878341771662, 0.396019809180871, 0.828948692418635, 0.608965877210721, 0.862616033758968, 0.511520224157721, 0.954354307847098, 0.537674984196201, 0.56923856260255, 0.631610019598156, 0.995451835216954, 0.706673043780029, 0.48861181223765, 0.0335634357761592, 0.260526915080845, 0.155798526015133, 0.538129250751808, 0.462861111620441, 0.832488533807918, 0.240286093903705, 0.804400400491431, 0.866366664646193, 0.491331875091419, 0.497079707449302, 0.00644026696681976, 0.264808948850259, 0.276109881000593, 0.151109413011, 0.624322594376281, 0.30089961993508, 0.785352566745132, 0.185590170091018, 0.275640194304287, 0.488332979613915, 0.691675447393209, 0.79187065269798, 0.977912291418761, 0.65804233122617, 0.496121389791369, 0.637790129054338, 0.354077852098271, 0.446584504563361, 0.0548810232430696, 0.977997847367078, 0.0446621906012297, 0.846142267808318, 0.792312033474445, 0.283239442389458, 0.569519791286439, 0.353674298850819, 0.873200476868078, 0.628311822423711, 0.092724812682718, 0.457397220889106, 0.610966128529981, 0.704369582701474, 0.096613202476874, 0.296744281193241, 0.606669862521812, 0.800369690405205, 0.530663885874674, 0.800844018813223, 0.694143302040175, 0.149135377025232, 0.574632864445448, 0.229900906328112, 0.850012331036851, 0.266677356557921, 0.119741068920121, 0.362008390016854, 0.57311961450614, 0.0860613696277142, 0.736391919665039, 0.694241075776517, 0.937062395270914, 0.264145026216283, 0.386534951860085, 0.544733077520505, 0.754919325001538, 0.503803891595453, 0.401726904092357, 0.0034493007697165, 0.486496781697497, 0.387490348657593, 0.280988613143563, 0.779526248341426, 0.124299887102097, 0.964687650091946, 0.299825241789222, 0.530165009433404, 0.211976165184751, 0.02977493731305, 0.397579323034734, 0.0155248397495598, 0.151030339067802, 0.207982781343162, 0.939318786840886, 0.102637642528862, 0.758516903733835, 0.109660369344056, 0.859466302907094, 0.0771992648951709, 0.285528805339709, 0.199461396317929, 0.465976278297603, 0.0754206536803395, 0.843383443076164, 0.34830015571788, 0.505159320775419, 0.757791465613991, 0.246772307436913, 0.945643378188834, 0.333399733761325, 0.537796301301569, 0.0504582962021232, 0.709704150911421, 0.661043661413714, 0.845065815607086, 0.937414651038125, 0.973442061571404, 0.618074327940121, 0.421558949165046, 0.461580977542326, 0.381967674940825, 0.900543630355969, 0.799630253110081, 0.577282109996304, 0.423818833660334, 0.625720134703442, 0.263384737772867, 0.961462962208316, 0.409270290052518, 0.970951860770583, 0.990634409477934, 0.2385896127671, 0.565137380734086, 0.546981194755062, 0.647274330956861, 0.519879736937582, 0.918819044483826, 0.404951939126477, 0.838488277047873, 0.0394472687039524, 0.434166448423639, 0.447053693933412, 0.3061246171128, 0.568001577397808, 0.466545169241726, 0.17175256786868, 0.944327270612121, 0.0916409485507756, 0.395825326675549, 0.274622379336506, 0.790226358920336, 0.550378059968352, 0.1967367536854, 0.027615929255262, 0.96021455200389, 0.59031739179045, 0.204236602410674, 0.62835328374058, 0.533675504149869, 0.694894760614261, 0.419009543024004, 0.331797570921481, 0.164566792547703, 0.387665628455579, 0.105281453579664, 0.0619033963885158, 0.49787598894909, 0.304480803199112, 0.674321013968438, 0.792639528866857, 0.815441026119515, 0.506663286359981, 0.360947646200657, 0.0467723300680518, 0.699251239188015, 0.805568242678419, 0.672241251217201, 0.673308677738532, 0.261712942505255, 0.824727525003254, 0.121759279863909, 0.363331084605306, 0.514416741440073, 0.561044790549204, 0.156923211412504, 0.810564022976905, 0.823299333686009, 0.80450737150386, 0.969820325030014, 0.984842949546874, 0.422898530960083, 0.0482544859405607, 0.914563844213262, 0.794556837296113, 0.181450214469805, 0.827487298985943, 0.670610515167937, 0.976752058370039, 0.15903457836248, 0.411284371512011, 0.950032961321995, 0.540905534056947, 0.0498545039445162, 0.541545539628714, 0.874443533364683, 0.575210688402876, 0.678419647505507, 0.234118766849861, 0.663440625881776, 0.689944268437102, 0.196472032926977, 0.701657786266878, 0.0900836251676083, 0.924485658295453, 0.946372972335666, 0.387061695102602, 0.0524065315257758, 0.403822224121541, 0.143173740012571, 0.538459418341517, 0.66730082873255, 0.543273411924019, 0.618338286411017, 0.395799603313208, 0.420506059424952, 0.526590002467856, 0.000323040410876274, 0.965873110806569, 0.378231909824535, 0.988238004501909, 0.861048805294558, 0.0671433575917035, 0.859857673291117, 0.94410000462085, 0.654062481131405, 0.964248648146167, 0.760174851631746, 0.0629169992171228, 0.784395115217194, 0.482730320654809, 0.32510017300956, 0.694570812629536, 0.287233001319692, 0.410376715473831, 0.970057932427153, 0.719467389164492, 0.326057652244344, 0.189590327441692, 0.924177415668964, 0.0954098391812295, 0.0756899723783135, 0.415992300957441, 0.311583701288328, 0.745843766722828, 0.779768801061437, 0.890066665830091, 0.383296261541545, 0.13056100718677, 0.25975012825802, 0.710162204690278, 0.98346358910203, 0.666211373638362, 0.89128101686947, 0.843808491248637, 0.277679994236678, 0.853456269484013, 0.866729144239798, 0.12823542766273, 0.695985280675814, 0.560508555267006, 0.38120332476683, 0.831820384133607, 0.478123567765579, 0.202240722952411, 0.777328006690368, 0.486161286011338, 0.175085100578144, 0.715836997143924, 0.369386761914939, 0.73872420238331, 0.153941362164915, 0.460651330184191, 0.640958406496793, 0.428959902841598, 0.387301699258387, 0.0395701695233583, 0.602446178440005, 0.627923281863332, 0.613305818056688, 0.41036047716625, 0.914007725426927, 0.382269617635757, 0.60592098115012, 0.396768094971776, 0.635859589092433, 0.609867009799927, 0.0201556966640055, 0.346673402702436, 0.178231648402289, 0.27840281277895, 0.863944157958031, 0.174572101561353, 0.913808126933873, 0.123761205701157, 0.770605004159734, 0.460414233617485, 0.544315242907032, 0.999929680489004, 0.74291770439595, 0.38524629524909, 0.744506446411833, 0.0539899193681777, 0.339884788263589, 0.588133748853579, 0.719801447121426, 0.79800266562961, 0.223058416042477, 0.387119446648285, 0.552028675097972, 0.825478129321709, 0.840955473249778, 0.187107299687341, 0.678104949183762, 0.2815468958579, 0.0968161327764392, 0.650165111292154, 0.29398794635199, 0.815197770018131, 0.196671800222248, 0.510698875878006, 0.546556799439713, 0.836086952360347, 0.538689216133207, 0.232996202772483, 0.460554061457515, 0.0290728490799665, 0.475463581038639, 0.540449279826134, 0.292340513318777, 0.255631507374346, 0.704852144233882, 0.121112965978682, 0.755290946923196, 0.76059159100987, 0.468732693698257, 0.44337248778902, 0.717496962286532, 0.818991089705378, 0.0846750373020768, 0.428250696742907, 0.337693152250722, 0.0131252487190068, 0.847056418657303, 0.768599247094244, 0.36215385212563, 0.657643290003762, 0.267831864766777, 0.598708418663591, 0.263055357849225, 0.722302006790414, 0.327617350965738, 0.93876527575776, 0.795239663217217, 0.434276887448505, 0.360423226607963, 0.459835171001032, 0.829221621388569, 0.70547627704218, 0.369496015598997, 0.558931387262419, 0.134506664937362, 0.904322380432859, 0.429697179701179, 0.46004637167789, 0.36806665873155, 0.134713468607515, 0.711025095079094, 0.0355947923380882, 0.621920493198559, 0.732832332840189, 0.542463185964152, 0.777340677566826, 0.653023166814819, 0.421225527301431, 0.0150399054400623, 0.219626130536199, 0.0825145454145968, 0.346807017922401, 0.047020765254274, 0.231877484126016, 0.180354218464345, 0.3627892464865, 0.848844629945233, 0.715603086166084, 0.347000322304666, 0.157886200584471, 0.757845731219277, 0.347851878032088, 0.108511561527848, 0.58932971348986, 0.682395635405555, 0.708306283922866, 0.421829289756715, 0.316923442762345, 0.706940085394308, 0.226845157565549, 0.706805334892124, 0.0214242502115667, 0.165778851602226, 0.866097510326654, 0.326134928967804, 0.820423212368041, 0.0926640175748616, 0.418068375671282, 0.278466330608353), actual = c(0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"