Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.2990851        0.2981578 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.3914156        0.3914156        0.4638615 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.293367855018005, 0.881186976563185, 0.556042284704745, 0.411801814101636, 0.549593400675803, 0.356222134083509, 0.459181251935661, 0.0689215890597552, 0.898389529436827, 0.168029899243265, 0.249229768058285, 0.899378034751862, 0.108827401185408, 0.839344532694668, 0.646934773307294, 0.611231360584497, 0.162534716539085, 0.408511965768412, 0.974364798981696, 0.988995987921953, 0.192241537617519, 0.898357023252174, 0.0300104550551623, 0.388066979125142, 0.372852022992447, 0.924334911163896, 0.211811531335115, 0.334602092392743, 0.339764140080661, 0.818414317676798, 0.835413446417078, 0.502736517461017, 0.751604082528502, 0.118741632672027, 0.616704010171816, 0.158164789900184, 0.0748787638731301, 0.226944346679375, 0.57507409597747, 0.476749832276255, 0.367602408165112, 0.0163039674516767, 0.928703480400145, 0.982911341823637, 0.442435957258567, 0.139678191859275, 0.907339117256925, 0.944059234112501, 0.304011464817449, 0.554833652451634, 0.704694246407598, 0.657702348195016, 0.655054502654821, 0.298754686955363, 0.192414895398542, 0.679476342629641, 0.203727402724326, 0.340274352114648, 0.33348248898983, 0.462511615129188, 0.80868262774311, 0.712768248049542, 0.109730609925464, 0.173798879608512, 0.0899656212422997, 0.457974986871704, 0.191939339740202, 0.192543046781793, 0.742693138774484, 0.869423697469756, 0.758233217056841, 0.762267811223865, 0.162845919607207, 0.439363436074927, 0.321110761491582, 0.01130959414877, 0.262209110427648, 0.0471166132483631, 0.507361272349954, 0.379007134120911, 0.0800735466182232, 0.198068294441327, 0.528551906580105, 0.993371254764497, 0.401151406345889, 0.163781315786764, 0.699102571234107, 0.0170565810985863, 0.39529606513679, 0.708933786256239, 0.252609489252791, 0.528763907030225, 0.402928872732446, 0.90045614191331, 0.0625389381311834, 0.668433183571324, 0.414764656219631, 0.186276175314561, 0.076383201405406, 0.596041114069521, 0.648630091687664, 0.145668484270573, 0.0865422438364476, 0.328539531910792, 0.0480204345658422, 0.074321563821286, 0.754424797138199, 0.793759891297668, 0.976533390348777, 0.772961619542912, 0.929710605647415, 0.537528971442953, 0.594125038012862, 0.587915395619348, 0.441798988264054, 0.0422495177481323, 0.774396046297625, 0.368184986058623, 0.175454635173082, 0.500654425472021, 0.0940139531157911, 0.650522663723677, 0.0290119764395058, 0.817227250663564, 0.200993573060259, 0.676805607043207, 0.186290014302358, 0.430816735140979, 0.277840813156217, 0.974294330691919, 0.362973711919039, 0.9042864064686, 0.758226859616116, 0.712546064285561, 0.1332922950387, 0.788233430124819, 0.544039955362678, 0.548782098805532, 0.303768766345456, 0.443727705162019, 0.971637289971113, 0.950677660526708, 0.82846062630415, 0.447388907661662, 0.207073725061491, 0.846311630681157, 0.263770162127912, 0.0193828211631626, 0.111059981165454, 0.631387222325429, 0.8687210441567, 0.385490261716768, 0.298817799659446, 0.633495466783643, 0.309178448980674, 0.770764025626704, 0.528478927211836, 0.257910927059129, 0.0400081551633775, 0.591983662219718, 0.652786374557763, 0.87504263757728, 0.915307695744559, 0.0287857092916965, 0.882237534737214, 0.581019898410887, 0.833912781672552, 0.234663509530947, 0.848127842647955, 0.921085855457932, 0.394928299589083, 0.342690402176231, 0.786513404920697, 0.304369719699025, 0.492842569248751, 0.0819860189221799, 0.349405946442857, 0.20667478418909, 0.648490600055084, 0.841317538172007, 0.225063255056739, 0.666405372787267, 0.343493655323982, 0.192435370758176, 0.689300769008696, 0.0576389359775931, 0.688832218991593, 0.0384172210469842, 0.73261943878606, 0.241522832307965, 0.994575324468315, 0.976397181861103, 0.435972415842116, 0.440256378147751, 0.936171762412414, 0.468665760010481, 0.442782649770379, 0.895463677123189, 0.746877709636465, 0.890374982263893, 0.498046248452738, 0.594223073450848, 0.183282458921894, 0.168402873445302, 0.670704191317782, 0.146796240936965, 0.945536506362259, 0.0192659676540643, 0.766265671467409, 0.95402133394964, 0.760511891683564, 0.0798832252621651, 0.396341816755012, 0.132929168641567, 0.708066290942952, 0.376133017241955, 0.174351781839505, 0.498672958696261, 0.62313349545002, 0.40290740900673, 0.115384308854118, 0.19041397399269, 0.577509038848802, 0.346814712509513, 0.474697291152552, 0.984569892985746, 0.110828183125705, 0.71420971211046, 0.695829121861607, 0.203896797727793, 0.240828937385231, 0.00231776363216341, 0.624308117898181, 0.554023018572479, 0.358639983693138, 0.658268300583586, 0.666608723346144, 0.956802225206047, 0.802923897048458, 0.747858471469954, 0.542369130300358, 0.19726982107386, 0.2468423305545, 0.345166991464794, 0.470452763373032, 0.477151662576944, 0.188550271093845, 0.956559848738834, 0.65486878156662, 0.178981061792001, 0.703436728799716, 0.320648925611749, 0.231414278270677, 0.764799266587943, 0.82607261906378, 0.327304515987635, 0.428188573569059, 0.367956734495237, 0.540377670666203, 0.343245230382308, 0.0361154023557901, 0.935399873880669, 0.931448213756084, 0.290883121546358, 0.814464873168617, 0.293647988932207, 0.205272305989638, 0.317099838517606, 0.0928299601655453, 0.021225348347798, 0.0943673131987453, 0.932629473740235, 0.0784541955217719, 0.576145481085405, 0.808143898146227, 0.818040573270991, 0.0372390833217651, 0.391538694035262, 0.300014914479107, 0.302176007069647, 0.134331145789474, 0.708311974070966, 0.849268369376659, 0.268293066881597, 0.548714301548898, 0.509326753905043, 0.794040007749572, 0.0519641565624624, 0.802013748791069, 0.221854542149231, 0.697842806577682, 0.974376777885482, 0.176604324253276, 0.75755439302884, 0.855028127552941, 0.399297064403072, 0.877252432284877, 0.507940975483507, 0.868104538181797, 0.753217927180231, 0.245570394909009, 0.625977728748694, 0.287493503419682, 0.590353612788022, 0.740422304952517, 0.771477302769199, 0.305720104370266, 0.702008564257994, 0.469234464224428, 0.0862213058862835, 0.0758664107415825, 0.574374636402354, 0.645343788899481, 0.652343594701961, 0.254693561699241, 0.08378088590689, 0.649108388461173, 0.873890327988192, 0.752188799902797, 0.585079514421523, 0.985025515314192, 0.345743647543713, 0.758030063938349, 0.654641238972545, 0.0402299435809255, 0.033586923032999, 0.453526697354391, 0.259968978352845, 0.248558914288878, 0.74229531432502, 0.848250419134274, 0.560173430712894, 0.514003506628796, 0.366410996066406, 0.534843531902879, 0.537830445449799, 0.586915302090347, 0.279286656528711, 0.659907097462565, 0.78965205932036, 0.0701472787186503, 0.521342914784327, 0.136883256258443, 0.762905292911455, 0.738885122118518, 0.229532372672111, 0.799980557290837, 0.697233381448314, 0.539603175362572, 0.836951052304357, 0.263221852248535, 0.204616746399552, 0.535433014156297, 0.853515729308128, 0.702833088114858, 0.111312191700563, 0.765025114640594, 0.312425981974229, 0.783514930866659, 0.541674505453557, 0.470168867148459, 0.763950430322438, 0.325212695403025, 0.693671178072691, 0.217925457749516, 0.997674428392202, 0.382925220765173, 0.338019811548293, 0.599581269314513, 0.321643013274297, 0.720844847150147, 0.326300173532218, 0.174297273159027, 0.0420233488548547, 0.521201073657721, 0.964739275164902, 0.734785261098295, 0.0986649948172271, 0.986433547921479, 0.179572999710217, 0.656291969120502, 0.346886682091281, 0.177801597164944, 0.591377912787721, 0.207867140648887, 0.326064110500738, 0.474989699199796, 0.59441485372372, 0.100470130331814, 0.0635067340917885, 0.672343876212835, 0.626752526499331, 0.672887949505821, 0.866352593991905, 0.662001677090302, 0.00263511762022972, 0.193193094106391, 0.603884334210306, 0.909007615642622, 0.620094646466896, 0.732201947364956, 0.545140663860366, 0.761072969995439, 0.588480619480833, 0.329447181895375, 0.976361198350787, 0.767241080757231, 0.287725572474301, 0.10940583772026, 0.211095143342391, 0.866986576700583, 0.0506005010101944, 0.614163890248165, 0.776625560130924, 0.774660362396389, 0.40014040353708, 0.838315196102485, 0.139260560506955, 0.0199095390271395, 0.199252681806684, 0.0529556442052126, 0.883697157725692, 0.983936433913186, 0.135641579516232, 0.286785591393709, 0.748073792783543, 0.105496160918847, 0.127358217490837, 0.822002782020718, 0.9911486832425, 0.0409127830062062, 0.237360443221405, 0.195050137815997, 0.574999026721343, 0.754288462921977, 0.463003684300929, 0.393552490742877, 0.282993301749229, 0.0640619741752744, 0.100965643767267, 0.970762035809457, 0.73035248205997, 0.565024686278775, 0.920414111576974, 0.515957348514348, 0.92946993210353, 0.317478775046766, 0.151094223139808, 0.849485176382586, 0.986302561592311, 0.556779641890898, 0.248722545104101, 0.50165510061197, 0.918152991915122, 0.44117276603356, 0.0705816310364753, 0.633361871354282, 0.129376601660624, 0.0453716251067817, 0.313119248487055, 0.443231935845688, 0.604014601092786, 0.0374939164612442, 0.364394396077842, 0.015712454682216, 0.828970142640173, 0.790572248399258, 0.134087256155908, 0.933299319352955, 0.312607677886263, 0.131239195587113, 0.678262818837538, 0.593728513456881, 0.0077829381916672, 0.750998901901767, 0.139604946598411, 0.8104449424427, 0.36066197254695, 0.773912098025903, 0.823199779726565, 0.657865367131308, 0.19732713396661, 0.187872296664864, 0.83940728683956, 0.219618956558406, 0.00566816562786698, 0.503661279100925, 0.60600006673485, 0.300157392164692, 0.0138388387858868, 0.349788460880518, 0.39097787020728, 0.7610999122262, 0.597374002449214, 0.564548382069916, 0.649874204304069, 0.246201655128971, 0.819828981766477, 0.464756367495283, 0.600214205449447, 0.878060930641368, 0.675418867496774, 0.0333504281006753, 0.702550526708364, 0.154111888259649, 0.00542453955858946, 0.350538900587708, 0.469983330927789, 0.607195073505864, 0.478905678493902, 0.522295564878732, 0.965047256788239, 0.599110534414649, 0.786984688835219, 0.0222244104370475, 0.157163607189432, 0.232004619902, 0.935619998257607, 0.671519361203536, 0.59810492163524, 0.179359741043299, 0.716291809920222, 0.171924542635679, 0.96135268942453, 0.991418252931908, 0.716584467794746, 0.561486776219681, 0.738297793781385, 0.563513694098219, 0.259987398050725, 0.779732874827459, 0.525506824022159, 0.888619243167341, 0.823416652157903, 0.0566701414063573, 0.169727986678481, 0.116075792815536, 0.00174234947189689, 0.410993659403175, 0.726235861424357, 0.188784453785047, 0.602236651349813, 0.387042117072269, 0.261055807117373, 0.867534605087712, 0.955698816338554, 0.454097833251581, 0.228573406348005, 0.551516306120902, 0.510118682635948, 0.596531430492178, 0.318373503163457, 0.981201893882826, 0.748626008164138, 0.30449889972806, 0.883019317872822, 0.845975951757282, 0.186333280988038, 0.943958092480898, 0.445780558977276, 0.00795683357864618, 0.932909255847335, 0.283186786808074, 0.12819932680577, 0.424823582638055, 0.760495719267055, 0.872693294193596, 0.125996533548459, 0.37906572339125, 0.77515078545548, 0.974860746413469, 0.123391105327755, 0.991214679088444, 0.935894710244611, 0.323608277365565, 0.532171761617064, 0.00831138831563294, 0.491656956961378, 0.0618765985127538, 0.10379069740884, 0.60238912794739, 0.033473230432719, 0.406480744481087, 0.449375936761498, 0.373982322169468, 0.862096894765273, 0.644548466894776, 0.886361971497536, 0.98647748795338, 0.430037733865902, 0.887559054186568, 0.183945128228515, 0.362892689881846, 0.0596177482511848, 0.132173098158091, 0.124551093671471, 0.679548022337258, 0.114197445102036, 0.273752615554258, 0.149741602363065, 0.694421417545527, 0.0560745561961085, 0.338976627215743, 0.771092545241117, 0.40275601670146, 0.526351391803473, 0.2554089166224, 0.980866901576519, 0.875485466560349, 0.51727109006606, 0.436085641384125, 0.792921896791086, 0.255257449578494, 0.932602355722338, 0.479517292696983, 0.683785810833797, 0.0194963435642421, 0.712542797671631, 0.764889722922817, 0.851276519009843, 0.849317326908931, 0.335634277202189, 0.651559178484604, 0.0233724645804614, 0.0411575012840331, 0.13645899342373, 0.630368244135752, 0.230067361146212, 0.392917728517205, 0.185437321662903, 0.701432762201875, 0.628080086782575, 0.293496093014255, 0.514284643344581, 0.744334376649931, 0.280015418073162, 0.310490436153486, 0.390721181407571, 0.851629330543801, 0.639067615149543, 0.330544350901619, 0.0971716593485326, 0.50425630249083, 0.80870395549573, 0.915240922011435, 0.0643262257799506, 0.583689791383222, 0.379863080801442, 0.00283461017534137, 0.0244002649560571, 0.30472527560778, 0.386750158620998, 0.325526443077251, 0.464831459801644, 0.64789793244563, 0.370265741366893, 0.988673516316339, 0.663287474075332, 0.77764740283601, 0.411267509451136, 0.427846495294943, 0.180632767733186, 0.830128917237744, 0.0677998585160822, 0.667616143124178, 0.132386727258563, 0.463325229473412, 0.559768125647679, 0.901446252595633, 0.169386293273419, 0.715376932406798, 0.570456452202052, 0.410933668958023, 0.0806601054500788, 0.423244249774143, 0.579522866290063, 0.630541944876313, 0.777284712763503, 0.0196927590295672, 0.463969753822312, 0.0047556352801621, 0.85437171603553, 0.162582563236356, 0.490053630899638, 0.669122624676675, 0.287526180502027, 0.142019507009536, 0.694731010822579, 0.211036340100691, 0.737673751078546, 0.935614126035944, 0.7945882815402, 0.856123517500237, 0.263258639723063, 0.51922318409197, 0.51354940305464, 0.585545218549669, 0.624510545050725, 0.58530356362462, 0.231058904668316, 0.748567885253578, 0.271512037608773, 0.796864505624399, 0.373894169693813, 0.26718248706311, 0.349467224441469, 0.526622390374541, 0.249266925733536, 0.820237761596218, 0.625968731939793, 0.69243388203904, 0.55524928518571, 0.458983677905053, 0.855945125222206, 0.0331878412980586, 0.0257637382019311, 0.279100898420438, 0.527317711850628, 0.750005944399163, 0.851465298328549, 0.819372311700135, 0.252097450429574, 0.241043630056083, 0.287226252723485, 0.445385552942753, 0.992335169110447, 0.690714735537767, 0.307554700179026, 0.174195643514395, 0.464321705745533, 0.689603109145537, 0.00636771251447499, 0.780254177981988, 0.614270880585536, 0.16787607036531, 0.359623890370131, 0.480199829675257, 0.867458332097158, 0.298618921544403, 0.647403566399589, 0.807350184535608, 0.0887870197184384, 0.577435258310288, 0.822783773299307, 0.409660895820707, 0.425562345422804, 0.607045772019774, 0.496134698623791, 0.358698622789234, 0.697897690115497, 0.950665555428714, 0.451645150547847, 0.0206703958101571, 0.0330052180215716, 0.953200537245721, 0.648564176866785, 0.965177268488333, 0.407599240308627, 0.834650815930218, 0.813597259577364, 0.439037949312478, 0.974069261923432, 0.646561581874266, 0.443553516408429, 0.453213336411864, 0.3270343686454, 0.394873419310898, 0.754885100759566, 0.473636634647846, 0.135816648602486, 0.110527332872152, 0.0809276674408466, 0.637412772048265, 0.142546548973769, 0.0934422980062664, 0.327816957840696, 0.934761571930721, 0.144292060984299, 0.763547992799431, 0.117001727921888, 0.779609415913001, 0.0915032331831753, 0.0823879707604647, 0.0298428165260702, 0.584968450013548, 0.585857881233096, 0.394156302791089, 0.75918801035732, 0.303523912094533, 0.621843648608774, 0.213477219222113, 0.400063798530027, 0.843356727622449, 0.165875074220821, 0.584621114656329, 0.368355576414615, 0.484074681531638, 0.415411515394226, 0.243681923020631, 0.868435412878171, 0.360448146704584, 0.259963082149625, 0.260078181978315, 0.5103037529625, 0.416885259561241, 0.313773645553738, 0.500033367890865, 0.90483810682781, 0.837411076528952, 0.642709798412398, 0.628366449149325, 0.368608248187229, 0.957687709480524, 0.352404982317239, 0.0431829954031855, 0.759980836184695, 0.563288539880887, 0.514713321113959, 0.553414875175804, 0.463081966619939, 0.682444753358141, 0.0225343350321054, 0.734496989287436, 0.453565900679678, 0.23073492012918, 0.126887693768367, 0.888548293616623, 0.72847977373749, 0.429179578321055, 0.742390793049708, 0.359224688028917, 0.853566232370213, 0.400139954639599, 0.0897187313530594, 0.295485098147765, 0.580551393795758, 0.615077819908038, 0.287099784938619, 0.615726323798299, 0.410712665878236, 0.075158150633797, 0.365580841433257, 0.817313212668523, 0.389689688803628, 0.748082171892747, 0.0286667041946203, 0.101490808883682, 0.893205165863037, 0.187140652211383, 0.640238580293953, 0.0215634333435446, 0.79397857375443, 0.785673661157489, 0.42740844306536, 0.660001650452614, 0.753806779393926, 0.27923946804367, 0.534162042662501, 0.485524835530668, 0.47274805791676, 0.185810704017058, 0.0796130457893014, 0.33442351478152, 0.694395089056343, 0.45369513402693, 0.608183174394071, 0.0870074052363634, 0.210463088471442, 0.857738120947033, 0.559121339814737, 0.15685576107353, 0.443964804057032, 0.307882418856025, 0.987615818157792, 0.826273276703432, 0.962545463582501, 0.448529122630134, 0.59835964650847, 0.621784107992426, 0.359194468706846, 0.751587339444086, 0.272161208791658, 0.32902890141122, 0.381872314726934, 0.153212159406394, 0.479331501759589, 0.0262957732193172, 0.229423091514036, 0.385848340345547, 0.644759890623391, 0.109849876724184, 0.438760028919205, 0.910281704505906, 0.0174366468563676, 0.479234723374248, 0.338921890361235, 0.067499402910471, 0.933149211807176, 0.730290068080649, 0.113651889609173, 0.256257302360609, 0.0589964224491268, 0.340458730934188, 0.893522829283029, 0.186974650481716, 0.175452890107408, 0.792257457971573, 0.454615780618042, 0.241291147191077, 0.422544350847602, 0.720931602874771, 0.326903771143407, 0.915108543122187, 0.000549150630831718, 0.993024228839204, 0.287581470096484, 0.240522884763777, 0.527120264945552, 0.514141545630991, 0.314976742491126, 0.411372743546963, 0.482587020844221, 0.369611728237942, 0.592385772615671, 0.878382567083463, 0.107940717134625, 0.869526305701584, 0.293125359574333, 0.499344053678215, 0.744433442130685, 0.679377225693315, 0.562098269350827, 0.69700739835389, 0.479418142465875, 0.503533739130944, 0.954771014163271, 0.162178047699854, 0.0274749717209488, 0.0339969508349895, 0.554915982298553, 0.28561293031089, 0.503130530007184, 0.0306190745905042, 0.656149522168562, 0.0995393383782357, 0.18631579563953, 0.676424997160211, 0.137616278138012, 0.431650279089808, 0.123017449630424, 0.887139515718445, 0.773322636028752, 0.997370916418731, 0.41537659522146, 0.778238249011338, 0.955603996291757, 0.927554095163941, 0.241373236756772, 0.143672083970159, 0.011781046167016, 0.829072596272454, 0.178065129090101, 0.0343529337551445, 0.846148843178526, 0.0455080934334546, 0.180059009930119, 0.717546085594222, 0.391611165367067, 0.47116130660288, 0.393199327634647, 0.504023938439786, 0.951822147937492, 0.875181755283847, 0.281849360326305, 0.90735610993579, 0.0285803617443889, 0.999652663944289, 0.421092183096334, 0.516176941106096, 0.604577445890754, 0.217728837858886, 0.130013276124373, 0.706648748368025, 0.183475801721215, 0.351195103023201, 0.556142598856241, 0.785513818496838, 0.410484571708366, 0.809224093332887, 0.526033774251118, 0.649266802472994, 0.426647383952513, 0.735948213376105, 0.670415641972795, 0.385423053056002, 0.118682828266174, 0.272050183033571, 0.432715818518773, 0.317119903629646, 0.566479028901085), actual = c(0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.618457046104595, 0.898061654064804, 0.267251870594919, 0.287810289533809, 0.657818253850564, 0.211743502877653, 0.0750168552622199, 0.798774918075651, 0.348186903866008, 0.847922842251137, 0.950096737826243, 0.868091696640477, 0.426432886626571, 0.991378817008808, 0.756815334316343, 0.0404631681740284, 0.0828251678030938, 0.260084298439324, 0.953951411647722, 0.271021711407229, 0.0451767584308982, 0.152809342835099, 0.533553258748725, 0.802307586185634, 0.67123704473488, 0.0411933928262442, 0.935130114434287, 0.287043133052066, 0.532045935513452, 0.94351054308936, 0.603873465210199, 0.422530974959955, 0.255963352974504, 0.60648883599788, 0.448996051447466, 0.0599568863399327, 0.836826708400622, 0.480939123779535, 0.0304182174149901, 0.513242230983451, 0.247082288609818, 0.958795955171809, 0.12430858775042, 0.949540361994877, 0.867936294525862, 0.946005751146004, 0.0820795777253807, 0.786885012406856, 0.20837211725302, 0.635472046677023, 0.636228266172111, 0.340152964927256, 0.747249078704044, 0.963224086677656, 0.293234422570094, 0.610766551923007, 0.647437882144004, 0.82640966353938, 0.436041216365993, 0.878458722494543, 0.030188858974725, 0.0136388095561415, 0.247149084461853, 0.767176839290187, 0.0530998453032225, 0.719615730224177, 0.612706919433549, 0.426295509096235, 0.965856689726934, 0.127168286824599, 0.50214695930481, 0.0982758810278028, 0.706960148643702, 0.633748127380386, 0.392331619048491, 0.261421816190705, 0.426943111466244, 0.369426198070869, 0.02319007483311, 0.441810854012147, 0.617089269449934, 0.526208368130028, 0.414609438041225, 0.014432908501476, 0.371889163041487, 0.654528341954574, 0.283916388638318, 0.821672202320769, 0.311215771129355, 0.770470175659284, 0.716921882238239, 0.0617180266417563, 0.296968641923741, 0.313537464477122, 0.969836382428184, 0.0961498853284866, 0.713378950022161, 0.911225982941687, 0.790615207282826, 0.694040755042806, 0.774646411417052, 0.676516831852496, 0.609338705893606, 0.545449123717844, 0.842782786116004, 0.244398277020082, 0.341399605385959, 0.22734494577162, 0.492015987168998, 0.746877451660112, 0.507770241703838, 0.153510042466223, 0.284998566843569, 0.117185020819306, 0.114602744812146, 0.329963782336563, 0.0859349798411131, 0.160787486704066, 0.851080609951168, 0.578772209584713, 0.812952000880614, 0.0196040936280042, 0.89411038835533, 0.311740816105157, 0.540705906925723, 0.35762034310028, 0.579944126075134, 0.170366988983005, 0.795447824755684, 0.196200930513442, 0.75363354687579, 0.447585865156725, 0.250750382430851, 0.476368553005159, 0.105277595343068, 0.350017891731113, 0.605254543013871, 0.815729767549783, 0.0455437863711268, 0.00560112204402685, 0.921732187503949, 0.177786808693781, 0.15583384456113, 0.0309512089006603, 0.501574632711709, 0.91256389953196, 0.497345547191799, 0.689371538348496, 0.918386537581682, 0.0677755910437554, 0.533294873777777, 0.467500020749867, 0.619764486793429, 0.761518852086738, 0.571556890383363, 0.536640637787059, 0.684971083654091, 0.429244248662144, 0.111429189797491, 0.930441341130063, 0.617047146894038, 0.507094274973497, 0.42238753195852, 0.184465323807672, 0.189957403577864, 0.349360833643004, 0.00567009043879807, 0.0615718686021864, 0.957758211065084, 0.202472453936934, 0.281796290073544, 0.754164298065007, 0.186373659875244, 0.185028420994058, 0.368197242962196, 0.149603624129668, 0.860406269319355, 0.51876090397127, 0.121456433553249, 0.981658346252516, 0.767416806425899, 0.0794930912088603, 0.0192945918533951, 0.634125303477049, 0.89921806496568, 0.234426420647651, 0.603762288345024, 0.241898840060458, 0.585908535169438, 0.381313037360087, 0.416141232708469, 0.565077397972345, 0.0864441622979939, 0.510958036174998, 0.975739562883973, 0.559512451291084, 0.959985956782475, 0.651921248994768, 0.718736793613061, 0.566569968359545, 0.716253861319274, 0.849982310319319, 0.10576139157638, 0.54519404284656, 0.803017216967419, 0.160903126001358, 0.564805636880919, 0.0804510226007551, 0.920867282664403, 0.411242653150111, 0.493913816520944, 0.580632185563445, 0.0956824359018356, 0.882807254325598, 0.043516629608348, 0.897964607691392, 0.875443747965619, 0.808137972373515, 0.451472618849948, 0.390148367034271, 0.325648821657524, 0.736139778513461, 0.350583590799943, 0.587761817732826, 0.815256858011708, 0.0876049157232046, 0.806215971475467, 0.456907899351791, 0.123651788337156, 0.426534430589527, 0.0434982085134834, 0.370437072822824, 0.782973546767607, 0.660139560932294, 0.539960565278307, 0.641437430167571, 0.408961987821385, 0.358730810927227, 0.902311836369336, 0.456849778071046, 0.705410664668307, 0.924876713659614, 0.974705803208053, 0.282191139645875, 0.534619391430169, 0.642623943276703, 0.367561713093892, 0.20618779072538, 0.549638699973002, 0.816692947642878, 0.485716886119917, 0.0250473318155855, 0.391299954382703, 0.204199058469385, 0.510582543211058, 0.244774354156107, 0.451061358908191, 0.634603671263903, 0.580300686880946, 0.821302657481283, 0.857482795603573, 0.713183178333566, 0.0284390742890537, 0.963653392158449, 0.66178297623992, 0.0333046626765281, 0.616649887058884, 0.945602132473141, 0.020344736520201, 0.929788895649835, 0.81787623395212, 0.87344227777794, 0.261043523205444, 0.0260872570797801, 0.680079785175622, 0.903565897839144, 0.418381063966081, 0.476640365552157, 0.207861400442198, 0.565862505231053, 0.120429839007556, 0.720206542173401, 0.90104355965741, 0.455589258577675, 0.416283077094704, 0.129561475710943, 0.175152689218521, 0.822142743505538, 0.869857589714229, 0.0733937851618975, 0.728753848699853, 0.98222532402724, 0.0442254713270813, 0.0864578818436712, 0.544191108550876, 0.864455543924123, 0.429026450728998, 0.0510361986234784, 0.56384359463118, 0.496822931803763, 0.460812111152336, 0.640551963355392, 0.792687831679359, 0.120672498364002, 0.2247387052048, 0.806729158386588, 0.377371783135459, 0.114956456935033, 0.0806798627600074, 0.840742012253031, 0.255218087462708, 0.221492245793343, 0.839728663442656, 0.459377835737541, 0.0478388834744692, 0.251050421269611, 0.645790772745386, 0.801829213276505, 0.411842286353931, 0.817116535501555, 0.564165007555857, 0.862498610746115, 0.34782120003365, 0.0601400493178517, 0.100593563169241, 0.930018780520186, 0.683938617119566, 0.729042568011209, 0.979756851214916, 0.850850640330464, 0.577340033138171, 0.972347869537771, 0.791754589881748, 0.375953377922997, 0.183639124734327, 0.867633337387815, 0.0372983976267278, 0.751848773797974, 0.45238332869485, 0.669503837358207, 0.227877579862252, 0.531176916323602, 0.973978184396401, 0.551593441748992, 0.914935666136444, 0.0258110677823424, 0.75637704692781, 0.194249479100108, 0.737921832129359, 0.896054191980511, 0.0162590658292174, 0.824254124658182, 0.598464567447081, 0.0816747304052114, 0.726133410120383, 0.596617000875995, 0.849970214534551, 0.722822174429893, 0.116090135416016, 0.540451289154589, 0.768128487747163, 0.58895693696104, 0.0658124361652881, 0.493961119093001, 0.136754979379475, 0.237920448184013, 0.397939581889659, 0.470000616740435, 0.530274881049991, 0.605416452046484, 0.513534131692722, 0.122773900395259, 0.746763237984851, 0.512958668870851, 0.0493417293764651, 0.806555225281045, 0.636246545938775, 0.608185109682381, 0.252925793174654, 0.208934204420075, 0.0572039056569338, 0.56870156340301, 0.17257842188701, 0.753977891756222, 0.14423620258458, 0.745944477152079, 0.477549495408311, 0.860718403244391, 0.65604772279039, 0.694229772780091, 0.519530097721145, 0.846512307878584, 0.861143801826984, 0.28674063552171, 0.230145090725273, 0.718516765162349, 0.822495885426179, 0.0408810768276453, 0.303234583698213, 0.70867947419174, 0.120273668318987, 0.964984395075589, 0.776310270885006, 0.685981201007962, 0.243578489171341, 0.0352118879090995, 0.0316841811873019, 0.0205503324978054, 0.0937494128011167, 0.124357930617407, 0.474851058796048, 0.638969569234177, 0.00744148925878108, 0.0509158368222415, 0.201618276536465, 0.140959108481184, 0.309893290977925, 0.976293220883235, 0.395316849462688, 0.136348160449415, 0.590774876298383, 0.812067597871646, 0.700022790813819, 0.908521945122629, 0.783289015991613, 0.983842957066372, 0.238390371436253, 0.901459624292329, 0.883836352499202, 0.470698258373886, 0.0218344849999994, 0.609707315918058, 0.822119887685403, 0.645347620360553, 0.196787333814427, 0.21498617483303, 0.792366802459583, 0.687843162333593, 0.226416183868423, 0.587699062889442, 0.607785024913028, 0.302097888663411, 0.408180992817506, 0.409233749145642, 0.646455829264596, 0.635738934390247, 0.202386408811435, 0.956895152805373, 0.180021652253345, 0.88392380415462, 0.851813934044912, 0.105674683582038, 0.0757078144233674, 0.908634688705206, 0.317487243795767, 0.466473842272535, 0.665796992834657, 0.692935814149678, 0.491283206734806, 0.5834652450867, 0.956917903851718, 0.98495588125661, 0.0435177066829056, 0.879966544918716, 0.801473940955475, 0.494375552982092, 0.786485988646746, 0.561678240541369, 0.295124516356736, 0.269806610653177, 0.400326968636364, 0.340763159561902, 0.198163368506357, 0.779705098364502, 0.28171267057769, 0.874984532827511, 0.351935042534024, 0.299218846485019, 0.532609866233543, 0.218289893819019, 0.896475994028151, 0.185614190530032, 0.694476437289268, 0.631428563268855, 0.0054243931081146, 0.448275322560221, 0.551308962982148, 0.938543494325131, 0.792495342437178, 0.468415785813704, 0.520793712465093, 0.923823330318555, 0.504979978548363, 0.093453595880419, 0.574435948161408, 0.467012469656765, 0.509620526107028, 0.429152509197593, 0.0039732726290822, 0.392585084307939, 0.430444571422413, 0.889186041429639, 0.200248998589814, 0.855571050895378, 0.815356702776626, 0.638173200655729, 0.891264582518488, 0.911301963264123, 0.240492133656517, 0.595711471280083, 0.544260926544666, 0.965459071099758, 0.589154336368665, 0.359748331597075, 0.98652866971679, 0.929015634348616, 0.0845738474745303, 0.00628080195747316, 0.105658045271412, 0.0860095941461623, 0.294576202752069, 0.357087813317776, 0.753395652398467, 0.0699205133132637, 0.693119299365208, 0.510172281647101, 0.89692568173632, 0.162703201407567, 0.84083480713889, 0.196927427779883, 0.290539343841374, 0.858095496427268, 0.532238724874333, 0.840933487052098, 0.949710857123137, 0.930994907626882, 0.478081656852737, 0.799151546088979, 0.246206167619675, 0.566664420301095, 0.601413427619264, 0.59302169480361, 0.641968481009826, 0.985017967643216, 0.621582899941131, 0.750697200186551, 0.540926629444584, 0.371955296956003, 0.847541031427681, 0.838481073034927, 0.170119860442355, 0.890363193582743, 0.467785774962977, 0.0176289977971464, 0.481677938252687, 0.172378222923726, 0.610724679194391, 0.262608028948307, 0.296967044938356, 0.770690062316135, 0.442404536297545, 0.437239169608802, 0.76655115140602, 0.52074020029977, 0.845858752029017, 0.902830263832584, 0.257101537892595, 0.717854872345924, 0.949497597292066, 0.542025451315567, 0.591136580798775, 0.443388113519177, 0.264427036978304, 0.862143334234133, 0.224209610838443, 0.0123903066851199, 0.0796306633856148, 0.497625693678856, 0.126735885627568, 0.784360353369266, 0.96949129155837, 0.230777422664687, 0.339526053750888, 0.35445539583452, 0.279899501940235, 0.747997479047626, 0.585516058374196, 0.881869382224977, 0.759910656139255, 0.580502803437412, 0.241157449083403, 0.451739177806303, 0.471979897236452, 0.0751900917384773, 0.0720896113198251, 0.389825235586613, 0.393337466288358, 0.856508533470333, 0.261503315530717, 0.535589740611613, 0.437839972320944, 0.443159565329552, 0.478161791805178, 0.0182476218324155, 0.3853540555574, 0.412653286010027, 0.902932873228565, 0.212207628414035, 0.421931131044403, 0.337312081828713, 0.129869013326243, 0.814669710118324, 0.0208549979142845, 0.651729031000286, 0.637334670173004, 0.975316789932549, 0.624499936122447, 0.273879193235189, 0.575275046750903, 0.374699575360864, 0.93069293955341, 0.405829900410026, 0.466052191099152, 0.436775120906532, 0.325269920052961, 0.579939382616431, 0.337254888610914, 0.127469090744853, 0.486617309274152, 0.230960050132126, 0.850006292806938, 0.341237290529534, 0.316999019822106, 0.878014136804268, 0.278450958663598, 0.58572341944091, 0.772935830056667, 0.638478421140462, 0.655803877161816, 0.158680712105706, 0.927223489386961, 0.0464006217662245, 0.546433693962172, 0.644877082435414, 0.521267913049087, 0.987429284024984, 0.976185931125656, 0.320575502933934, 0.367710778256878, 0.427567828679457, 0.0347765262704343, 0.998684705235064, 0.602184449788183, 0.724091843701899, 0.745689862873405, 0.437296542106196, 0.854646014980972, 0.161925460910425, 0.661066617351025, 0.361667001619935, 0.641930781072006, 0.128692717757076, 0.371426684549078, 0.454432338941842, 0.157925982959569, 0.180736460955814, 0.97420370974578, 0.0996429766528308, 0.930982374586165, 0.851628688629717, 0.0490661924704909, 0.849176211748272, 0.222525791963562, 0.894101016223431, 0.198898026021197, 0.492952710948884, 0.638465869240463, 0.654072847450152, 0.821155748562887, 0.131413507508114, 0.249434241792187, 0.293367739068344, 0.130903784418479, 0.330478814197704, 0.010248088510707, 0.553616914199665, 0.704026692081243, 0.0600850605405867, 0.453910564305261, 0.109671354992315, 0.302896868437529, 0.525316102197394, 0.428401978919283, 0.941262997686863, 0.0264094814192504, 0.618816832778975, 0.236019420437515, 0.0161691894754767, 0.522625962272286, 0.835880485363305, 0.707511955406517, 0.506610169541091, 0.923958163475618, 0.370830476982519, 0.374257132643834, 0.749465356580913, 0.296613338403404, 0.40185624663718, 0.314103108365089, 0.956156411906704, 0.257444738410413, 0.309140803525224, 0.141823927173391, 0.257164113456383, 0.107640143483877, 0.900068979244679, 0.839285149704665, 0.652147905668244, 0.856953764567152, 0.545026443433017, 0.0891570223029703, 0.427564420737326, 0.316438545938581, 0.486213606549427, 0.0462820352986455, 0.671642928617075, 0.943078660871834, 0.749812448630109, 0.76763425883837, 0.460388766834512, 0.829888813197613, 0.948072675149888, 0.583902457496151, 0.138168275821954, 0.590087026823312, 0.0894202422350645, 0.693634690716863, 0.0988893832545727, 0.109498021192849, 0.232114198384807, 0.895783408312127, 0.246695182751864, 0.131033814745024, 0.403700375929475, 0.23884057207033, 0.0741053221281618, 0.950279884506017, 0.190932906931266, 0.149020169861615, 0.00422668713144958, 0.0923979000654072, 0.862225338816643, 0.573853831971064, 0.531719534890726, 0.807993356604129, 0.928861761698499, 0.98668677918613, 0.230538316071033, 0.935409356141463, 0.0710469980258495, 0.759374147979543, 0.988518938189372, 0.582935333484784, 0.994507383555174, 0.847298691980541, 0.195839478401467, 0.511074282694608, 0.312424393137917, 0.278562132501975, 0.730023265816271, 0.0662396603729576, 0.0518018156290054, 0.166046979138628, 0.572243800852448, 0.778079569572583, 0.646916340338066, 0.839295415207744, 0.969431994948536, 0.12885746662505, 0.0837568254210055, 0.732544824946672, 0.614693851675838, 0.159432735294104, 0.646773441461846, 0.199888050556183, 0.269679660676047, 0.277663832064718, 0.625256122322753, 0.431320126634091, 0.658699993975461, 0.196912535931915, 0.0222111567854881, 0.251031872117892, 0.674333871575072, 0.112861709902063, 0.453634567791596, 0.549471677746624, 0.986160592408851, 0.0850787961389869, 0.160179572645575, 0.314495040802285, 0.24836144875735, 0.872409189585596, 0.90304914326407, 0.0792234321124852, 0.223616964416578, 0.784642922692001, 0.287950274068862, 0.239840158959851, 0.220031301723793, 0.606570420321077, 0.474948097486049, 0.43309706565924, 0.0929573674220592, 0.296265708049759, 0.509622826008126, 0.96993566560559, 0.490711068734527, 0.891998453764245, 0.520344782155007, 0.544137248536572, 0.34244257514365, 0.285615648375824, 0.215206935768947, 0.945266759255901, 0.828641435131431, 0.700416354229674, 0.853105068672448, 0.905785501468927, 0.263542779488489, 0.0524967722594738, 0.201081014471129, 0.569855675334111, 0.576640625484288, 0.807265242328867, 0.901868998073041, 0.00853819889016449, 0.881231712875888, 0.913055590353906, 0.14220308791846, 0.798196669435129, 0.731872532749549, 0.723776148399338, 0.630187263013795, 0.325616865651682, 0.204494869336486, 0.884718678891659, 0.452854811446741, 0.795328004285693, 0.635160895064473, 0.184325780486688, 0.462760205380619, 0.34262441075407, 0.393279203446582, 0.583511709934101, 0.0685383428353816, 0.575037089409307, 0.561584409559146, 0.411323700100183, 0.598441487411037, 0.377788475248963, 0.837374445283785, 0.705547713208944, 0.0268831378780305, 0.431218462996185, 0.539266819367185, 0.294190537650138, 0.250331598101184, 0.148774383123964, 0.596283551072702, 0.604570549447089, 0.190364087466151, 0.266737198224291, 0.806498101213947, 0.530201309360564, 0.52965625631623, 0.662797597236931, 0.371280990540981, 0.919795733876526, 0.720337870530784, 0.60439092409797, 0.384383941767737, 0.330382201820612, 0.244216242106631, 0.522813575575128, 0.364429080160335, 0.36470294254832, 0.106209565652534, 0.131712871603668, 0.991499055176973, 0.586155772674829, 0.168607394909486, 0.803068250417709, 0.0774589066859335, 0.108917265431955, 0.0404290251899511, 0.932441040873528, 0.0266940265428275, 0.966736480593681, 0.512730269925669, 0.27948680985719, 0.682981327641755, 0.428844112204388, 0.223364857723936, 0.320957048097625, 0.0222682130988687, 0.807217339519411, 0.721884652273729, 0.667231640312821, 0.222215032670647, 0.3735720384866, 0.621761160669848, 0.460011078743264, 0.889905874617398, 0.493715052260086, 0.379597953520715, 0.190811720211059, 0.708568573463708, 0.671651575015858, 0.475377040682361, 0.493452072376385, 0.171028999844566, 0.25604199920781, 0.347635967424139, 0.794189515523612, 0.286315225064754, 0.440646375762299, 0.775923616718501, 0.372958539286628, 0.474430707748979, 0.0641984140966088, 0.0276455050334334, 0.258913847617805, 0.30689738993533, 0.627035881159827, 0.819334001280367, 0.914026395650581, 0.00456715654581785, 0.94011010043323, 0.161421157419682, 0.626739178784192, 0.947516256244853, 0.594400276662782, 0.522215215023607, 0.322821092791855, 0.501087274868041, 0.366146100685, 0.857532755238935, 0.658445546869189, 0.539558114483953, 0.337018173187971, 0.643806607695296, 0.148142484016716, 0.797946312464774, 0.944832334062085, 0.137179754208773, 0.997487719170749, 0.801313889445737, 0.820243204478174, 0.586581105366349, 0.530822487547994, 0.105431789299473, 0.689584015170112, 0.300969709875062, 0.199135644594207, 0.711004640208557, 0.134304794948548, 0.814958985662088, 0.281547093531117, 0.0193176385946572, 0.948325234698132, 0.00264561641961336, 0.314590637106448, 0.922096184222028, 0.206489804666489, 0.894971884088591, 0.637467612279579, 0.856980922399089, 0.198516510659829, 0.538736071903259, 0.776551947230473, 0.490409842459485, 0.125372339971364, 0.256541722686961, 0.0411485165823251, 0.872223421698436, 0.217555239796638, 0.654208173509687, 0.367674650158733, 0.274867894593626, 0.032797136111185, 0.529865224380046, 0.146194382570684, 0.0214719718787819, 0.329600146505982, 0.791482037631795, 0.115853191353381, 0.156062226509675, 0.774675778811797, 0.0298303961753845), actual = c(0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"