Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4856107        0.4856107 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4856107        0.4856107        0.4678749 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.611944541567937, 0.913510862737894, 0.344150326913223, 0.223307891283184, 0.372574327280745, 0.120204949053004, 0.471162548521534, 0.223002668702975, 0.0262803130317479, 0.736905645346269, 0.461785145103931, 0.310447811847553, 0.143712132936344, 0.0668934811837971, 0.351130313239992, 0.508378091966733, 0.771638736827299, 0.718401754042134, 0.581768821924925, 0.743779249489307, 0.245401488151401, 0.932778432965279, 0.654950215714052, 0.979560282081366, 0.190305596916005, 0.346847621491179, 0.434643030865118, 0.748579526785761, 0.87302535912022, 0.579827542183921, 0.588165508583188, 0.843348813476041, 0.924255674704909, 0.198285881895572, 0.370919591281563, 0.293473955709487, 0.190346721326932, 0.807018654653803, 0.975583229213953, 0.538678326644003, 0.4143878640607, 0.184646956855431, 0.779074307763949, 0.94844379182905, 0.753116248873994, 0.00460413796827197, 0.616180654615164, 0.210875391494483, 0.370919407345355, 0.0113321428652853, 0.332423752406612, 0.1546902845148, 0.22978695272468, 0.115511557552963, 0.0326603017747402, 0.183655854081735, 0.128028759965673, 0.606526921037585, 0.262642018729821, 0.163911270676181, 0.334651520242915, 0.626761034829542, 0.159290131414309, 0.967076934175566, 0.414071714505553, 0.140789062017575, 0.189301784150302, 0.139858451671898, 0.0772110775578767, 0.676411447813734, 0.0844120974652469, 0.476290793623775, 0.0998171546962112, 0.232348386896774, 0.952658590395004, 0.195079104974866, 0.935187906958163, 0.218363885767758, 0.612349343718961, 0.695789628429338, 0.332246815552935, 0.295112204970792, 0.710097316652536, 0.72387043456547, 0.898280111374334, 0.253638375550508, 0.790322215063497, 0.407178180525079, 0.189968796912581, 0.96906975004822, 0.492646786384284, 0.625594586599618, 0.741868268931285, 0.790371873183176, 0.867031435947865, 0.130896838847548, 0.881885478971526, 0.942725895904005, 0.312727509532124, 0.466174808098003, 0.588922271504998, 0.532710891216993, 0.148748867446557, 0.728956964565441, 0.222659912193194, 0.0719390590675175, 0.0828415811993182, 0.835260790307075, 0.0149020589888096, 0.52253388799727, 0.702864034799859, 0.862903142580763, 0.257720390101895, 0.986811595968902, 0.0678966853301972, 0.38991260365583, 0.804374874569476, 0.545411640545353, 0.632078010356054, 0.12382174259983, 0.301693363580853, 0.271630090195686, 0.315693284384906, 0.599435698008165, 0.538843551417813, 0.752532629994676, 0.20594407687895, 0.916145693976432, 0.327215261757374, 0.464650296373293, 0.988836487755179, 0.408171216258779, 0.965426495531574, 0.691822342108935, 0.0619887893553823, 0.335332968039438, 0.263241452397779, 0.334790482418612, 0.165662793908268, 0.642231982201338, 0.00601862394250929, 0.463698236038908, 0.614138294709846, 0.702285960782319, 0.306089702295139, 0.513642768142745, 0.404563536401838, 0.707231582840905, 0.316656451206654, 0.494176246691495, 0.534741110634059, 0.144308463670313, 0.0561576739419252, 0.702033440815285, 0.164392887847498, 0.655180925969034, 0.958083639387041, 0.275383887579665, 0.332836321322247, 0.957517593167722, 0.929968155920506, 0.615124604431912, 0.320762211922556, 0.248000327730551, 0.558033983455971, 0.396559104323387, 0.999524229671806, 0.862939173355699, 0.389441792620346, 0.712110011605546, 0.906699111685157, 0.195482524111867, 0.023964875144884, 0.368348993360996, 0.213227240601555, 0.842035160399973, 0.382278041681275, 0.418321022298187, 0.244053112342954, 0.537341542541981, 0.784465520875528, 0.756537700304762, 0.559709401801229, 0.469414979917929, 0.157848480856046, 0.461145469220355, 0.715864554047585, 0.856860842788592, 0.821115448372439, 0.483410025481135, 0.647559118224308, 0.578003652626649, 0.759823093889281, 0.882322695571929, 0.664367797318846, 0.0464261658489704, 0.273659773403779, 0.840246552368626, 0.212153460597619, 0.195996631169692, 0.31892817793414, 0.63351554912515, 0.765932405367494, 0.428483141819015, 0.964846815681085, 0.00811124173924327, 0.408974946476519, 0.854874543612823, 0.111059315269813, 0.698962445138022, 0.53047079150565, 0.675902869086713, 0.266120453830808, 0.542966938344762, 0.646986378589645, 0.0378782250918448, 0.655584338353947, 0.336781225632876, 0.0566765489056706, 0.395952842198312, 0.596119297202677, 0.394750977633521, 0.258581655565649, 0.491918093059212, 0.216399824479595, 0.202543589984998, 0.110824275761843, 0.407165463315323, 0.288851137738675, 0.847837331937626, 0.758152381749824, 0.235868567833677, 0.203033117111772, 0.91256599733606, 0.84643363696523, 0.81563963717781, 0.618975925026461, 0.38458179612644, 0.277856009313837, 0.85570504842326, 0.939568073023111, 0.556716567836702, 0.339303502812982, 0.933274274226278, 0.842189322225749, 0.849693802883849, 0.662162326741964, 0.322654193732888, 0.329941744916141, 0.606234605889767, 0.894112287089229, 0.796500320080668, 0.9015891070012, 0.636800950393081, 0.996547660790384, 0.921959623461589, 0.409795791376382, 0.531703704036772, 0.115824956912547, 0.610369951697066, 0.262940475018695, 0.419600148685277, 0.235893631353974, 0.72312474460341, 0.257659132825211, 0.495743680512533, 0.0413695210590959, 0.503545831423253, 0.336001113988459, 0.107994557823986, 0.4479739300441, 0.868493074318394, 0.232457245932892, 0.446497110184282, 0.542754755355418, 0.779538402799517, 0.489440443459898, 0.70974186854437, 0.259100128896534, 0.72079806169495, 0.207877102773637, 0.498869855422527, 0.341759458649904, 0.387433651601896, 0.299541804241017, 0.279162406921387, 0.391508348053321, 0.278767026495188, 0.403944663703442, 0.977008031448349, 0.245896042790264, 0.763948077103123, 0.586451241979375, 0.810668866615742, 0.33560694148764, 0.917227673344314, 0.317894092295319, 0.37382367416285, 0.760654440149665, 0.915473674656823, 0.825212991097942, 0.735399243887514, 0.579551531467587, 0.435137226246297, 0.192906012525782, 0.0809621149674058, 0.970067433314398, 0.678070536581799, 0.0892184311524034, 0.74894234072417, 0.940798111725599, 0.747637256514281, 0.539673069259152, 0.0293180851731449, 0.793653879780322, 0.833610596135259, 0.122956591425464, 0.267497346037999, 0.0151894360315055, 0.02101338095963, 0.373874795390293, 0.802338553592563, 0.209140992490575, 0.338609780417755, 0.68648669985123, 0.0461291067767888, 0.605823863297701, 0.836251865606755, 0.0560624343343079, 0.608042991254479, 0.986094889231026, 0.0576933154370636, 0.433479934232309, 0.603591111022979, 0.875231855316088, 0.597385715460405, 0.983071177499369, 0.400739529170096, 0.778223136439919, 0.703826177632436, 0.166909974766895, 0.184410441666842, 0.922286005457863, 0.393094989238307, 0.470206155208871, 0.128273031441495, 0.0237310100346804, 0.551170227117836, 0.294563833856955, 0.793740101857111, 0.0779808543156832, 0.906677619088441, 0.501444971188903, 0.620142568368465, 0.0316149848513305, 0.391074838815257, 0.13939298177138, 0.220837211003527, 0.672307487344369, 0.551733244908974, 0.781780946766958, 0.861557440832257, 0.405396985355765, 0.0229718426708132, 0.0737372641451657, 0.586299131624401, 0.55455209990032, 0.0861099979374558, 0.607854368630797, 0.0348422052338719, 0.564414295367897, 0.113689398858696, 0.0109098094981164, 0.949957689037547, 0.358819389715791, 0.477503573987633, 0.948055661981925, 0.749750379705802, 0.877455542096868, 0.745800625300035, 0.613645823905244, 0.921279879286885, 0.377519755857065, 0.32063638442196, 0.358247167663649, 0.47550271009095, 0.798418477643281, 0.00161101552657783, 0.492058644071221, 0.879895469173789, 0.533319121226668, 0.750095603056252, 0.209984756773338, 0.931514978408813, 0.803711923770607, 0.586755552561954, 0.722759593976662, 0.301432769279927, 0.44488611840643, 0.805406751576811, 0.32863924652338, 0.582023052033037, 0.0298205681610852, 0.960296459263191, 0.815873846411705, 0.197564522502944, 0.109377839369699, 0.871465682052076, 0.513588468078524, 0.370265319477767, 0.380489233415574, 0.312185555230826, 0.374994568759575, 0.650291118072346, 0.18088404671289, 0.624552099499851, 0.486699731787667, 0.268442565342411, 0.453025605063885, 0.451449268264696, 0.0792656494304538, 0.260050134733319, 0.556177521822974, 0.716431165812537, 0.247933767037466, 0.390497406944633, 0.853425033157691, 0.482217214303091, 0.300701993284747, 0.558848829008639, 0.806260090088472, 0.26083259168081, 0.121846783207729, 0.371416875161231, 0.197560806991532, 0.733476695138961, 0.108589518116787, 0.199725350830704, 0.726145764114335, 0.543352146167308, 0.597114552976564, 0.237848317949101, 0.415429158834741, 0.407466486794874, 0.561929850373417, 0.56022772542201, 0.764534313697368, 0.329658834030852, 0.230993524426594, 0.0730384441558272, 0.793464120710269, 0.653805539477617, 0.437568960012868, 0.421711598988622, 0.566446192795411, 0.657009015558288, 0.047264012740925, 0.0525817496236414, 0.236488857073709, 0.0349840265698731, 0.213775259675458, 0.317409583600238, 0.215436578495428, 0.752333186333999, 0.112805559532717, 0.838875090470538, 0.790051368065178, 0.011069749481976, 0.415666723623872, 0.574255769141018, 0.0853075904306024, 0.810206056339666, 0.962655475363135, 0.249801684636623, 0.943683901568875, 0.0132973005529493, 0.796909951604903, 0.803584795910865, 0.631430210079998, 0.652691350551322, 0.0829636189155281, 0.325727263232693, 0.809102937811986, 0.376240774057806, 0.726552697829902, 0.449288142379373, 0.413023649016395, 0.515685832826421, 0.89897361327894, 0.628194143762812, 0.210206424118951, 0.50048236688599, 0.547347160521895, 0.654988055350259, 0.831521259853616, 0.488798203179613, 0.966917598154396, 0.721373594598845, 0.260593090206385, 0.199683752376586, 0.868381215026602, 0.59909657202661, 0.901974630309269, 0.757873490219936, 0.85238300007768, 0.350835786899552, 0.530861605657265, 0.488641422474757, 0.629725868580863, 0.893705178750679, 0.52538445382379, 0.982914303662255, 0.595568995457143, 0.0740295031573623, 0.507020665332675, 0.328597155399621, 0.408092852914706, 0.908980366541073, 0.985312583856285, 0.616480068070814, 0.0871233609504998, 0.0739937711041421, 0.654821997741237, 0.229832426412031, 0.755670476239175, 0.41397015703842, 0.96799634816125, 0.558539457852021, 0.76983999623917, 0.632106591714546, 0.110872007207945, 0.537305665900931, 0.996643433813006, 0.843607705552131, 0.816738222260028, 0.232797649456188, 0.815200794721022, 0.152315428713337, 0.266734568635002, 0.686220507137477, 0.181585424812511, 0.00691661331802607, 0.179619470611215, 0.802054683212191, 0.58656041440554, 0.846172189572826, 0.441303388914093, 0.0987136529292911, 0.0901824692264199, 0.82837457023561, 0.612793822772801, 0.106020615436137, 0.288298336556181, 0.615933377295732, 0.59948970284313, 0.370874262880534, 0.287798649864271, 0.516180722508579, 0.460045042447746, 0.239358686609194, 0.269541095709428, 0.652233843458816, 0.561897762352601, 0.809122205479071, 0.680961893172935, 0.105566764250398, 0.841228175675496, 0.381573268445209, 0.645815508672968, 0.105942626250908, 0.371460132766515, 0.238595630042255, 0.195965671213344, 0.219799007521942, 0.760679414262995, 0.323695028200746, 0.423184149432927, 0.12683611153625, 0.284000411164016, 0.863803245825693, 0.379993550013751, 0.817424326436594, 0.85978595004417, 0.930006200447679, 0.864850013749674, 0.011737153865397, 0.951910377014428, 0.255422002403066, 0.933194737415761, 0.634702567942441, 0.285686698043719, 0.502067333320156, 0.440041762776673, 0.994349020300433, 0.892239174572751, 0.0118647075723857, 0.789823122788221, 0.492637141840532, 0.18168924911879, 0.274748124415055, 0.65870580682531, 0.964174226624891, 0.96364592644386, 0.379120633471757, 0.831084294477478, 0.646274618804455, 0.782652478432283, 0.982668936485425, 0.348723290488124, 0.606420510914177, 0.574200505157933, 0.451786546735093, 0.902651981450617, 0.283487556967884, 0.887363938847557, 0.860395669005811, 0.394066525623202, 0.710043603787199, 0.383332318393514, 0.49791890499182, 0.36060756794177, 0.186722569866106, 0.758117107441649, 0.727096359943971, 0.889311102218926, 0.865097331115976, 0.262395930010825, 0.432618904625997, 0.127331585856155, 0.248168357182294, 0.327299550641328, 0.753566364990547, 0.4676588405855, 0.9670841882471, 0.43195455125533, 0.136770761338994, 0.0581892372574657, 0.46968441689387, 0.543326931772754, 0.867299066390842, 0.274277856806293, 0.317458448698744, 0.383535183267668, 0.961413599317893, 0.557070306967944, 0.349316248903051, 0.823397730709985, 0.124697125982493, 0.587500249734148, 0.496373420348391, 0.96458137081936, 0.917300362838432, 0.407535957172513, 0.83877713070251, 0.0959701396059245, 0.452696998137981, 0.885523601667956, 0.471862695645541, 0.224806740181521, 0.0188938016071916, 0.605594580061734, 0.199407980777323, 0.0828353858087212, 0.995242406614125, 0.189540426013991, 0.393567503197119, 0.996530645526946, 0.410520675126463, 0.816124318866059, 0.720271332655102, 0.672643191181123, 0.118089517345652, 0.230334390653297, 0.676811040146276, 0.382768931332976, 0.608302792068571, 0.0552242218982428, 0.623005707981065, 0.308611335931346, 0.82184256054461, 0.0670427745208144, 0.589807437965646, 0.752476847963408, 0.386172051308677, 0.463031101739034, 0.434955073287711, 0.235078875208274, 0.502404224127531, 0.101701900130138, 0.881774879060686, 0.902912479592487, 0.500213505700231, 0.93520286004059, 0.580610081553459, 0.599809407256544, 0.580943410750479, 0.576194396940991, 0.0698515246622264, 0.284162218682468, 0.211033292813227, 0.00602781632915139, 0.00294363335706294, 0.189026300795376, 0.686155806761235, 0.531721752835438, 0.313406188040972, 0.00672682351432741, 0.0446779318153858, 0.240028045838699, 0.302258540643379, 0.955624962691218, 0.72448764485307, 0.693527988158166, 0.167771139880642, 0.302203446626663, 0.713303476339206, 0.169994658092037, 0.941057159332559, 0.578927947441116, 0.762254068627954, 0.0748577064368874, 0.0150156908202916, 0.286916184704751, 0.167720062891021, 0.102651489200071, 0.489148022606969, 0.866375201381743, 0.337123492266983, 0.80619089351967, 0.875525737414137, 0.508717973250896, 0.548114036908373, 0.498931175097823, 0.95562193589285, 0.429356405045837, 0.665199656272307, 0.667509728344157, 0.301434258464724, 0.161281341919675, 0.0700687898788601, 0.349376130849123, 0.527488253545016, 0.283296116162091, 0.0976502529811114, 0.192877020454034, 0.389987800968811, 0.527433711802587, 0.138408295344561, 0.836949629476294, 0.75728490576148, 0.330263949232176, 0.403229519026354, 0.423812981694937, 0.349040642147884, 0.388788142241538, 0.88511119922623, 0.210846518399194, 0.471508368151262, 0.311677382560447, 0.90882891183719, 0.523812078172341, 0.936467528576031, 0.927148155635223, 0.0275936352554709, 0.15417081839405, 0.501648945501074, 0.390277395956218, 0.156529112253338, 0.544448427390307, 0.525010874029249, 0.387712315889075, 0.380241045262665, 0.299109001178294, 0.142450302373618, 0.934137009549886, 0.0588534472044557, 0.314550868235528, 0.446581515716389, 0.604643388884142, 0.605526073137298, 0.27039234363474, 0.17262067925185, 0.0157091447617859, 0.943980322917923, 0.856151354266331, 0.80383506254293, 0.301332165719941, 0.162618833128363, 0.885734460083768, 0.493648934876546, 0.519563071895391, 0.0229807121213526, 0.178557513514534, 0.676408528350294, 0.535136386053637, 0.159049437381327, 0.488169355783612, 0.142486400203779, 0.602177526801825, 0.575913422973827, 0.42863285029307, 0.563869819976389, 0.755335249006748, 0.0550720011815429, 0.578099889447913, 0.114728862652555, 0.78579908143729, 0.292051269672811, 0.437857149867341, 0.0416574301198125, 0.634067470673472, 0.0485040484927595, 0.828988942783326, 0.320749789709225, 0.300664235372096, 0.267517659114674, 0.908826331375167, 0.733707368839532, 0.790187386563048, 0.972260250709951, 0.488637787289917, 0.141701692482457, 0.10922977887094, 0.13113852054812, 0.824053576216102, 0.992431512800977, 0.623323632869869, 0.749453521100804, 0.853914967738092, 0.104944116668776, 0.394279272994027, 0.562381788156927, 0.275446158833802, 0.665733252651989, 0.403888889122754, 0.201496312860399, 0.771106104133651, 0.480604436714202, 0.029910346493125, 0.950313323410228, 0.319836334791034, 0.935119131579995, 0.573181690182537, 0.938233972527087, 0.024656584020704, 0.518923120107502, 0.406383702997118, 0.602745555574074, 0.416263580555096, 0.487914497265592, 0.792587887262926, 0.920480325352401, 0.887112540425733, 0.371470933314413, 0.882428422104567, 0.392151355277747, 0.0599992298521101, 0.352926509687677, 0.313013724982738, 0.997319592628628, 0.365702538518235, 0.988345603225753, 0.443970050197095, 0.435161089058965, 0.886595640098676, 0.0203605769202113, 0.731910649454221, 0.342432073084638, 0.544903619913384, 0.891583716496825, 0.0582430486101657, 0.164086386794224, 0.579070742242038, 0.238889129599556, 0.366515358909965, 0.925186990061775, 0.246969280764461, 0.113955527311191, 0.73480269103311, 0.994532074313611, 0.0875458915252239, 0.258324234513566, 0.154205305268988, 0.5845065433532, 0.580926687922329, 0.617483042646199, 0.918541637714952, 0.628889359766617, 0.204674893058836, 0.405752551509067, 0.891456762328744, 0.465750824427232, 0.284284113906324, 0.595522318733856, 0.563768201041967, 0.0665291727054864, 0.262791350716725, 0.0292474895250052, 0.607126155402511, 0.607321460964158, 0.311659859027714, 0.780624739825726, 0.367486510192975, 0.234000528696924, 0.917765001067892, 0.280161659000441, 0.823940463596955, 0.289554015733302, 0.927559900097549, 0.053115552989766, 0.411910229362547, 0.323713156860322, 0.901402844116092, 0.900935735320672, 0.548933474114165, 0.970886246999726, 0.998465505195782, 0.370525966165587, 0.975333668058738, 0.146635806653649, 0.0453183937352151, 0.177867505466565, 0.970109665300697, 0.369691952830181, 0.0593478886876255, 0.496284909313545, 0.521971521666273, 0.0873524101916701, 0.175268626539037, 0.388469873927534, 0.142882420681417, 0.972183370031416, 0.654936506412923, 0.808038893388584, 0.703220618655905, 0.92152234306559, 0.410943139810115, 0.44358038995415, 0.759642906021327, 0.75788554456085, 0.342113589867949, 0.428883755812421, 0.471438233740628, 0.57100494694896, 0.485714863287285, 0.245366809191182, 0.0277458755299449, 0.980636480497196, 0.218701508594677, 0.0231058846693486, 0.750858504325151, 0.550437367521226, 0.971690792124718, 0.451034662779421, 0.445239767665043, 0.686665758490562, 0.875916169723496, 0.0568623875733465, 0.0747788867447525, 0.986965706339106, 0.909655791707337, 0.266107017174363, 0.970204157289118, 0.275732695823535, 0.865464079426602, 0.124341744463891, 0.898205855861306, 0.817778618307784, 0.798968734452501, 0.239210764411837, 0.621539667714387, 0.736669374397025, 0.531674630474299, 0.457424694439396, 0.325981779722497, 0.157164985081181, 0.585321829421446, 0.888268236536533, 0.543541415594518, 0.552816027542576, 0.386895787203684, 0.9635580542963, 0.367971116676927, 0.783932606223971, 0.182566017610952, 0.459922882961109, 0.993707802146673, 0.812569512752816, 0.45409702998586, 0.901329695945606, 0.81421506148763, 0.0902902171947062, 0.917159980395809, 0.508735275128856, 0.187624324578792, 0.322749176295474, 0.353565685916692, 0.560451717115939, 0.929373743478209, 0.430384703911841), actual = c(1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.386394367553294, 0.099167064530775, 0.264972462318838, 0.094239275669679, 0.145095506217331, 0.749400848057121, 0.783974372781813, 0.137178664794192, 0.666803538100794, 0.000564364017918706, 0.000864079920575023, 0.275354271521792, 0.157598005374894, 0.592037634458393, 0.324001970468089, 0.0453947642818093, 0.772566627943888, 0.864271006314084, 0.653963355347514, 0.964154465822503, 0.716134130489081, 0.781519418116659, 0.48500332958065, 0.241240530274808, 0.616876928834245, 0.221126655349508, 0.0387130605522543, 0.182610098039731, 0.0861329182516783, 0.350776573875919, 0.426023534033448, 0.825445415684953, 0.171607898781076, 0.740153098711744, 0.154394047567621, 0.506973675219342, 0.96382588497363, 0.788072317605838, 0.917223235359415, 0.704132551094517, 0.60177919222042, 0.547508177347481, 0.380410116165876, 0.336626509437338, 0.884357813047245, 0.704379332484677, 0.498705435544252, 0.451211074599996, 0.538132184417918, 0.908210914116353, 0.573401278117672, 0.618465996114537, 0.166875606402755, 0.173084988025948, 0.940941203385592, 0.939448490971699, 0.48569256439805, 0.626392447156832, 0.861220686230808, 0.925641970243305, 0.0759828907903284, 0.975898965029046, 0.875358834397048, 0.102088470477611, 0.953691087197512, 0.466667610220611, 0.709299966227263, 0.668758888961747, 0.474634836427867, 0.313309432240203, 0.782953136833385, 0.558468005619943, 0.258018019609153, 0.857285151956603, 0.750972503330559, 0.0172921472694725, 0.510807670885697, 0.00448282202705741, 0.432471078587696, 0.803964659338817, 0.0102352746762335, 0.780192750971764, 0.871421553194523, 0.596532654250041, 0.898154536727816, 0.630695485277101, 0.418638124829158, 0.665376524440944, 0.0498274185229093, 0.328582911519334, 0.602810287848115, 0.775712711270899, 0.926323466701433, 0.44654752057977, 0.000820302870124578, 0.724909285781905, 0.277895962353796, 0.421136655611917, 0.0299235840793699, 0.610926266293973, 0.264699310529977, 0.358714305795729, 0.986002118792385, 0.895879694260657, 0.00104138813912868, 0.284461624687538, 0.417810707353055, 0.347363068256527, 0.0387109010480344, 0.949344954686239, 0.798844769131392, 0.352498154854402, 0.617243349552155, 0.792909283889458, 0.640592355048284, 0.879971004789695, 0.491670927032828, 0.404844128061086, 0.764877215027809, 0.915639809565619, 0.730794309172779, 0.115083448123187, 0.967021462740377, 0.594578848453239, 0.782115128356963, 0.442902802489698, 0.865343691781163, 0.103692578151822, 0.376787221059203, 0.921002448769286, 0.0448140422813594, 0.939668336883187, 0.282983961747959, 0.371085312217474, 0.234658434288576, 0.490384852513671, 0.698151305550709, 0.693772637750953, 0.466309025650844, 0.0195711657870561, 0.435128087177873, 0.520600261632353, 0.622704304521903, 0.34024469438009, 0.794473445508629, 0.392806758405641, 0.920973523054272, 0.388055577408522, 0.615226227557287, 0.192395110614598, 0.590649085119367, 0.585651482921094, 0.631926885340363, 0.86802395991981, 0.349707915913314, 0.7276195588056, 0.957178041338921, 0.4493091723416, 0.316429971251637, 0.59588117711246, 0.905468516750261, 0.261258692480624, 0.820816916180775, 0.370874142507091, 0.0345929234754294, 0.521838884335011, 0.407447905978188, 0.789536767173558, 0.720030914060771, 0.138320874655619, 0.342680507106707, 0.506291714496911, 0.579707237659022, 0.937843155814335, 0.773587162140757, 0.742582325823605, 0.855809459229931, 0.54690342140384, 0.645003751153126, 0.620497041847557, 0.960408275015652, 0.764695814810693, 0.0127912182360888, 0.470970283728093, 0.52391705987975, 0.08470807550475, 0.416440206579864, 0.19348461786285, 0.367101611569524, 0.216945285676047, 0.572359364712611, 0.258308667456731, 0.801752471365035, 0.231672327965498, 0.721957413945347, 0.831930227344856, 0.332866103388369, 0.659354384755716, 0.845477359369397, 0.775991527363658, 0.523913566488773, 0.297542835585773, 0.754020757740363, 0.647444281028584, 0.456858284305781, 0.0310114393942058, 0.703907215036452, 0.361882058670744, 0.925298655405641, 0.181434906786308, 0.541864663828164, 0.46217887615785, 0.691177945118397, 0.746355849551037, 0.960913151036948, 0.122775761410594, 0.515462992712855, 0.753305461024866, 0.97444678330794, 0.149268937762827, 0.499934462131932, 0.207251934101805, 0.0045031642075628, 0.189868373563513, 0.468410854460672, 0.229141740128398, 0.316262999316677, 0.298988419817761, 0.127306861802936, 0.608133286004886, 0.876831780886278, 0.78273876872845, 0.363213359611109, 0.948991580400616, 0.552678555250168, 0.129956599092111, 0.668717632768676, 0.533163217594847, 0.0824886320624501, 0.452766895759851, 0.392054900759831, 0.383451748173684, 0.713767167646438, 0.0154736314434558, 0.58930767676793, 0.967645198339596, 0.25639749923721, 0.696821901248768, 0.958639639196917, 0.989305652910843, 0.2796351313591, 0.546216696267948, 0.83065602183342, 0.450026265112683, 0.0480703974608332, 0.100111991167068, 0.616283369250596, 0.665985296247527, 0.7863923702389, 0.899585528764874, 0.677708144998178, 0.513713026419282, 0.778000194113702, 0.713896855013445, 0.452482952969149, 0.252146102720872, 0.0415352315176278, 0.44004163891077, 0.595220698276535, 0.637084898538888, 0.126857259776443, 0.388445425778627, 0.782779817702249, 0.429215816548094, 0.468209586339071, 0.841464485274628, 0.399255669210106, 0.710292222909629, 0.702277851756662, 0.0245880011934787, 0.779290019301698, 0.830534545704722, 0.0941695207729936, 0.606447979575023, 0.43905235058628, 0.65214012702927, 0.525486440397799, 0.0179902731906623, 0.195779223460704, 0.885371634969488, 0.953499621944502, 0.850789453135803, 0.551163957221434, 0.000979756470769644, 0.563610990764573, 0.695943207945675, 0.375509311445057, 0.354380434611812, 0.9509421242401, 0.102417483460158, 0.130151036893949, 0.566360239638016, 0.0179675440303981, 0.757824977394193, 0.576025301124901, 0.212633374845609, 0.631505845813081, 0.679208477959037, 0.357691588113084, 0.767440223600715, 0.582711689407006, 0.715778660494834, 0.216340311337262, 0.470805132295936, 0.603847309714183, 0.90979940444231, 0.119258700637147, 0.23488768003881, 0.711475231451914, 0.901337079470977, 0.426867839647457, 0.567895967746153, 0.170553537551314, 0.964842298999429, 0.732779500773177, 0.272847843822092, 0.441548571223393, 0.696227301843464, 0.48091091029346, 0.964317915728316, 0.0282172861043364, 0.786013810662553, 0.321602435549721, 0.369337921496481, 0.524618059862405, 0.84276467282325, 0.437657230533659, 0.108508042758331, 0.152128933230415, 0.25091337133199, 0.550343566341326, 0.634063704172149, 0.0891952840611339, 0.82759129232727, 0.614353082375601, 0.189681791700423, 0.767224300187081, 0.205982085084543, 0.0749342837370932, 0.56352688325569, 0.186862633563578, 0.662307479651645, 0.40462622535415, 0.722840971779078, 0.535525719868019, 0.433666258817539, 0.814210356678814, 0.56333679612726, 0.840107796480879, 0.903165282681584, 0.968563401605934, 0.94860578677617, 0.349234280176461, 0.703343820059672, 0.962719885865226, 0.699444571044296, 0.600461647612974, 0.555505373515189, 0.291803704574704, 0.797398218652233, 0.465185245964676, 0.548604312585667, 0.485942845931277, 0.589287295006216, 0.821804842445999, 0.287354608299211, 0.918079440714791, 0.210325305582955, 0.0988924163393676, 0.289019729243591, 0.463879461400211, 0.541404238436371, 0.515511643607169, 0.645925211021677, 0.373114459682256, 0.607728631235659, 0.633477268740535, 0.125565936323255, 0.167185449507087, 0.830349081661552, 0.796278206864372, 0.331052346155047, 0.126100983470678, 0.0325820315629244, 0.741924953879789, 0.573534508934245, 0.737686394248158, 0.658462717663497, 0.548692199168727, 0.752690055640414, 0.0513371485285461, 0.727316414471716, 0.365990335121751, 0.421151144662872, 0.252273489022627, 0.0308227110654116, 0.0296560698188841, 0.977163962321356, 0.899826030246913, 0.811451399698853, 0.36362387612462, 0.0635727997869253, 0.180978033691645, 0.40675984043628, 0.198220973834395, 0.442032055929303, 0.867883428465575, 0.531200196361169, 0.599217277718708, 0.269875919213519, 0.331423575058579, 0.62818236136809, 0.11658755550161, 0.790511371567845, 0.400965747190639, 0.770864381687716, 0.493464362574741, 0.152415959397331, 0.142330795293674, 0.03715042443946, 0.113600920652971, 0.343343608081341, 0.68023598426953, 0.850513909012079, 0.212739186361432, 0.257783950306475, 0.0523290429264307, 0.325167751871049, 0.0145057938061655, 0.562554859090596, 0.282382166711614, 0.910726664587855, 0.914822597522289, 0.981927424902096, 0.618229198502377, 0.889732132432982, 0.808433515019715, 0.4023109190166, 0.246306594926864, 0.670496995793656, 0.808388470904902, 0.651891287881881, 0.0482768754009157, 0.772140822606161, 0.631313152611256, 0.576149506028742, 0.309197419090196, 0.315222894307226, 0.154764718841761, 0.192951216362417, 0.597022437024862, 0.871401580516249, 0.968686468200758, 0.388094682944939, 0.555737797170877, 0.360559085151181, 0.791055661160499, 0.825633141910657, 0.0506674388889223, 0.671197491697967, 0.832411219365895, 0.037258229451254, 0.779084800044075, 0.873432551976293, 0.347417544340715, 0.074465268291533, 0.986403475981206, 0.0883079979103059, 0.525371973868459, 0.98354080086574, 0.31315298913978, 0.606639742152765, 0.873500946909189, 0.389822476310655, 0.424904136219993, 0.445451717590913, 0.52666758466512, 0.659232736565173, 0.47748756268993, 0.986852976027876, 0.982172964373603, 0.102624658262357, 0.226228864863515, 0.898433518595994, 0.3285522533115, 0.760664193658158, 0.966767210513353, 0.878290250431746, 0.313368866220117, 0.966174758272246, 0.787476825993508, 0.863185737282038, 0.922194336773828, 0.183008204214275, 0.724393710261211, 0.61492399405688, 0.91155225946568, 0.486923207528889, 0.313561040209606, 0.26435004430823, 0.0991611671634018, 0.872888391604647, 0.214601707179099, 0.872403831453994, 0.540669592795894, 0.332272000843659, 0.97484096349217, 0.760047789197415, 0.662012221058831, 0.0607198851648718, 0.791196261532605, 0.518664356088266, 0.293432794511318, 0.51102073607035, 0.828847153112292, 0.404727927874774, 0.380829994333908, 0.786313344258815, 0.997708245413378, 0.98699290654622, 0.883767875377089, 0.322638169396669, 0.784296578494832, 0.628401128109545, 0.780243715504184, 0.982166541041806, 0.466434188885614, 0.0394986516330391, 0.956660822732374, 0.36092264787294, 0.879239383386448, 0.0262012556195259, 0.55367405153811, 0.629446730483323, 0.635337178828195, 0.136906316503882, 0.821374242659658, 0.314628791529685, 0.979002771899104, 0.647114272462204, 0.572083455277607, 0.949460870586336, 0.844724491704255, 0.311933546094224, 0.784860692918301, 0.975505930604413, 0.465964763425291, 0.312811054289341, 0.436800960917026, 0.0945005593821406, 0.873822923749685, 0.0871596592478454, 0.507095664739609, 0.76665549329482, 0.155708993785083, 0.534937290940434, 0.380142816342413, 0.0297345996368676, 0.790835930965841, 0.582641669316217, 0.805782279232517, 0.519377053482458, 0.697847066447139, 0.194447985151783, 0.365871984511614, 0.411514491308481, 0.782617277931422, 0.848772945813835, 0.0761368139646947, 0.305942248785868, 0.747090723831207, 0.719502341700718, 0.621974061941728, 0.0795877987984568, 0.341426288941875, 0.815843408694491, 0.341477927751839, 0.940153343603015, 0.68143904581666, 0.449549736687914, 0.557718117954209, 0.394481600029394, 0.296568477991968, 0.437878219643608, 0.396228737197816, 0.047269486123696, 0.177273951470852, 0.287650476675481, 0.519430665532127, 0.679370132740587, 0.884258277015761, 0.0815327949821949, 0.267245343420655, 0.0704292333684862, 0.651431727223098, 0.765986929414794, 0.703898114385083, 0.49012984498404, 0.403406427241862, 0.510081707732752, 0.504240016918629, 0.804527402389795, 0.634577655466273, 0.582012346945703, 0.353213587310165, 0.214014108525589, 0.171915961196646, 0.409655367489904, 0.819945520022884, 0.146507471101359, 0.749722976237535, 0.138315035263076, 0.326825720025226, 0.0034021467436105, 0.828078964725137, 0.659173525171354, 0.846699664834887, 0.749624425778165, 0.0523508791811764, 0.147785968147218, 0.917504168348387, 0.103318691253662, 0.681051451945677, 0.841903411783278, 0.00226062443107367, 0.286143228877336, 0.0386590317357332, 0.528104827273637, 0.919485509395599, 0.273528554476798, 0.939404438948259, 0.228243184275925, 0.969553143018857, 0.0250596224796027, 0.0939564302098006, 0.23974565952085, 0.460039385827258, 0.549163563875481, 0.893828236497939, 0.569911096477881, 0.211043776711449, 0.386753401486203, 0.644218133529648, 0.100919430842623, 0.92587862117216, 0.0267811422236264, 0.147500513354316, 0.0815491562243551, 0.127777958521619, 0.128091315040365, 0.685384743381292, 0.715779312187806, 0.102420765440911, 0.22191175725311, 0.473845034604892, 0.494503671769053, 0.188799017574638, 0.374030618462712, 0.718899021390826, 0.219363438198343, 0.414742125431076, 0.541370584163815, 0.0384251717478037, 0.950715825892985, 0.85792481363751, 0.080736847827211, 0.138880504760891, 0.156368493335322, 0.262492368230596, 0.952909602085128, 0.0335413592401892, 0.972885519033298, 0.8979775859043, 0.130189448129386, 0.943674143636599, 0.0953877116553485, 0.970985035412014, 0.10556246060878, 0.875517505919561, 0.826138142729178, 0.489744118880481, 0.569051343016326, 0.538131090113893, 0.718719473807141, 0.113906855462119, 0.859784984728321, 0.226696701720357, 0.695720190415159, 0.0749808689579368, 0.0313431723043323, 0.165542524075136, 0.0916475506965071, 0.625729040941224, 0.591229945188388, 0.231593114789575, 0.468618632759899, 0.189861976774409, 0.620636600302532, 0.0260182681959122, 0.877276593819261, 0.798648584634066, 0.563398866215721, 0.705131095601246, 0.620947160525247, 0.0072003323584795, 0.942950023571029, 0.619507019408047, 0.35746070696041, 0.518813834525645, 0.510661868844181, 0.965511076152325, 0.155911706387997, 0.503161034779623, 0.753237294498831, 0.346287541789934, 0.445511289406568, 0.989025291986763, 0.135016532614827, 0.784384787315503, 0.872744465945289, 0.344030802836642, 0.125717166578397, 0.187609654152766, 0.766950721619651, 0.377442554803565, 0.557958225021139, 0.615215197205544, 0.318731558509171, 0.132995211984962, 0.677183405961841, 0.843197713140398, 0.473877302370965, 0.871865497203544, 0.622114450903609, 0.5420437741559, 0.909273604396731, 0.623778105946258, 0.147379898931831, 0.895790819311514, 0.0324639827013016, 0.734043953707442, 0.414899899624288, 0.111083779716864, 0.187385004479438, 0.671461104415357, 0.888388958759606, 0.12394139287062, 0.728544770041481, 0.740464788628742, 0.626891673309729, 0.683821636019275, 0.958573660813272, 0.85929575166665, 0.0777324046939611, 0.898290170123801, 0.44909533392638, 0.441206435440108, 0.333232863340527, 0.787605901947245, 0.336787920212373, 0.643911675317213, 0.832291705766693, 0.253950539976358, 0.581282896455377, 0.0759280039928854, 0.106070668203756, 0.793566850014031, 0.44893876905553, 0.141408673254773, 0.400216510519385, 0.41710976161994, 0.586823852732778, 0.435649126302451, 0.332223306410015, 0.08331074914895, 0.549144334625453, 0.355058365035802, 0.929345610784367, 0.210546042537317, 0.163922431413084, 0.102581051178277, 0.239310255041346, 0.08011574158445, 0.713602072792128, 0.0254080616869032, 0.521908048074692, 0.847511250292882, 0.148177582304925, 0.729483237490058, 0.134199291234836, 0.657456299290061, 0.145666875410825, 0.455549482721835, 0.798381904140115, 0.383478031028062, 0.161755438661203, 0.292589295655489, 0.163284153910354, 0.292022877372801, 0.776218129554763, 0.828284278744832, 0.212121451273561, 0.221558094490319, 0.365907567786053, 0.312325942562893, 0.330116491764784, 0.0655395549256355, 0.407912960276008, 0.79895158438012, 0.835705714300275, 0.479529552394524, 0.428794642211869, 0.386573669034988, 0.741386130452156, 0.833250233437866, 0.752440709387884, 0.673400769941509, 0.145946654491127, 0.353354471502826, 0.697120137512684, 0.357747671194375, 0.907102231169119, 0.84477694099769, 0.0508992313407362, 0.580539365764707, 0.583029630593956, 0.297790785087273, 0.773423589067534, 0.299449251033366, 0.759894823888317, 0.670634352602065, 0.389094598591328, 0.91868025274016, 0.239215760724619, 0.688732111360878, 0.0607830509543419, 0.347500823670998, 0.743201020173728, 0.59716519434005, 0.0101390171330422, 0.593711315421388, 0.363560564117506, 0.124871046515182, 0.441060272976756, 0.167525013443083, 0.807883489411324, 0.978320970898494, 0.25276743946597, 0.527494983747602, 0.578937422484159, 0.145380658330396, 0.302215479547158, 0.240734429564327, 0.754082349129021, 0.599759878357872, 0.815292809624225, 0.504988443106413, 0.744309430709109, 0.38703494425863, 0.750748763093725, 0.476415134500712, 0.768394393147901, 0.157473522238433, 0.235151716275141, 0.799348993692547, 0.480573820881546, 0.316448190016672, 0.505466291448101, 0.313566190889105, 0.791592210298404, 0.112342283362523, 0.291441740235314, 0.229998696828261, 0.575263338396326, 0.735091325128451, 0.628986361902207, 0.277810354484245, 0.00243738200515509, 0.815256523666903, 0.238072882872075, 0.560696326196194, 0.701138534350321, 0.0241665143985301, 0.464482716284692, 0.812226309906691, 0.674761610105634, 0.515932964161038, 0.414172392105684, 0.598886376479641, 0.403050650842488, 0.172383286291733, 0.478111712727696, 0.585451394552365, 0.478777762269601, 0.0983593668788671, 0.134687096579, 0.119463972281665, 0.0912195201963186, 0.447639414807782, 0.264468990731984, 0.327098864596337, 0.533251991262659, 0.615765723399818, 0.201139618409798, 0.270055024651811, 0.603701496031135, 0.723215804435313, 0.408310562837869, 0.305500411661342, 0.24202453577891, 0.349502632394433, 0.933610733598471, 0.758681459119543, 0.282378717325628, 0.239623139612377, 0.852800270542502, 0.694654791150242, 0.137084018206224, 0.35316473362036, 0.852640589000657, 0.152793003013358, 0.60983035620302, 0.48877679137513, 0.604542102431878, 0.472111684503034, 0.802638402907178, 0.00672856182791293, 0.979601545259356, 0.6552300369367, 0.127700916258618, 0.0499667534604669, 0.646632970077917, 0.844294583657756, 0.360814645187929, 0.648591449484229, 0.353667656658217, 0.772303594974801, 0.864201868884265, 0.289542808430269, 0.897186416434124, 0.600102481199428, 0.888504842296243, 0.901798099046573, 0.245961957145482, 0.353619999717921, 0.584128162125126, 0.15104608447291, 0.56123903230764, 0.235333399381489, 0.84552869736217, 0.0878644562326372, 0.218655046308413, 0.0236498680897057, 0.814156852662563, 0.646756879985332, 0.349079694366083, 0.766280876239762, 0.683404797222465, 0.556299490854144, 0.0804175173398107, 0.0693253935314715, 0.407894426956773, 0.680611342890188, 0.718148395419121, 0.385290952865034, 0.125514163635671, 0.798019701149315, 0.390844865469262, 0.496031064540148, 0.709050576901063, 0.393809138098732, 0.121618394739926, 0.681569418404251, 0.773721958277747, 0.135678020073101, 0.0334398683626205, 0.860305950511247, 0.754847807111219, 0.764961528358981, 0.1206384296529, 0.600686223246157, 0.875226734904572, 0.422450754558668, 0.984636922599748, 0.857330530649051, 0.0795986130833626, 0.0355195652227849, 0.76639010431245, 0.0598582159727812, 0.813080531312153), actual = c(1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"