Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.1427671        0.2008542 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4901224        0.4901224        0.5068011 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.64235448022373, 0.785124326823279, 0.356974031077698, 0.459815388079733, 0.781282065901905, 0.748638838762417, 0.246919760946184, 0.997096197213978, 0.171789491781965, 0.455663586501032, 0.0902166469022632, 0.959926962852478, 0.193157458445057, 0.252770679770038, 0.981529865181074, 0.471954925684258, 0.922323755454272, 0.908166498411447, 0.452265985542908, 0.912682235240936, 0.771939032711089, 0.250788378762081, 0.839713295223191, 0.411504688905552, 0.63080970197916, 0.854861823143438, 0.252894906094298, 0.738872425165027, 0.383994666393846, 0.222624376649037, 0.493368009570986, 0.693424870958552, 0.124468884430826, 0.60076120798476, 0.162916892208159, 0.372594031272456, 0.689206487964839, 0.7357718416024, 0.923993293428794, 0.737006694078445, 0.592629296472296, 0.904608338605613, 0.96095346705988, 0.925293101929128, 0.938423885963857, 0.377001620130613, 0.930635564727709, 0.416347302030772, 0.511117007350549, 0.730606195051223, 0.771089707501233, 0.242047832347453, 0.824307686882094, 0.0562227924820036, 0.809672859264538, 0.228849342325702, 0.304337591631338, 0.759054383030161, 0.276175621896982, 0.628921675495803, 0.168220325605944, 0.202041869750246, 0.543464802671224, 0.965635945089161, 0.176531908102334, 0.932374851545319, 0.637022576993331, 0.298313201637939, 0.48006292199716, 0.0443563640583307, 0.480397921288386, 0.249575063586235, 0.851227779872715, 0.857928533339873, 0.533822673605755, 0.914038760121912, 0.306111578829587, 0.851317962631583, 0.611403414048254, 0.323068500962108, 0.302601587725803, 0.440542545868084, 0.84444558667019, 0.154387026559561, 0.91334838164039, 0.245252368506044, 0.76165860821493, 0.228323714109138, 0.929895594483241, 0.15049013029784, 0.217775912256911, 0.367107865866274, 0.390770488651469, 0.168884907616302, 0.470511509338394, 0.361917595146224, 0.0414087721146643, 0.571874149842188, 0.650883801048622, 0.619349046843126, 0.719369416357949, 0.265696654561907, 0.00642164354212582, 0.877162611344829, 0.325376988854259, 0.257679603761062, 0.549014512216672, 0.127777588320896, 0.875514768064022, 0.87309808563441, 0.930616312893108, 0.888363096863031, 0.218733476940542, 0.136325398692861, 0.421139987884089, 0.972712855786085, 0.922521344386041, 0.863418448716402, 0.524191541830078, 0.177322418661788, 0.341025751549751, 0.736163033638149, 0.954464056529105, 0.012490798952058, 0.119422087911516, 0.915390803711489, 0.984553175279871, 0.413812499493361, 0.17960387095809, 0.113763091852888, 0.399966299999505, 0.908712618984282, 0.911372170550749, 0.369155266555026, 0.685318218311295, 0.79981507663615, 0.418562202015892, 0.607882176991552, 0.54477676586248, 0.0260521315503865, 0.761256035650149, 0.838537055067718, 0.0376649859827012, 0.0333015543874353, 0.335376675473526, 0.483436549082398, 0.452692501246929, 0.712268699891865, 0.855482122628018, 0.891103600151837, 0.80252012074925, 0.921096583595499, 0.995478352997452, 0.0914201072882861, 0.768780600745231, 0.263122564414516, 0.320963796461001, 0.294134467374533, 0.777887862175703, 0.770701214205474, 0.316438496811315, 0.328165922313929, 0.587140317074955, 0.185894950293005, 0.314159654546529, 0.884715050924569, 0.326472979737446, 0.746622662292793, 0.326956212054938, 0.00128167518414557, 0.927656502695754, 0.882520637009293, 0.669448742177337, 0.570458279224113, 0.501676662126556, 0.760778895812109, 0.83375585032627, 0.620947194751352, 0.321887822588906, 0.440715389559045, 0.631555017549545, 0.73703810153529, 0.00237917108461261, 0.822155788773671, 0.362469125306234, 0.963907072786242, 0.171539064031094, 0.124084464972839, 0.699298925464973, 0.752288496820256, 0.443541050422937, 0.31366282585077, 0.493099077837542, 0.956292609451339, 0.55446440144442, 0.40674304543063, 0.36004895856604, 0.261720321606845, 0.540237727807835, 0.0434316222090274, 0.0231908762361854, 0.024148614378646, 0.781617408618331, 0.851601066067815, 0.399858149001375, 0.585517220664769, 0.172283269930631, 0.870689766481519, 0.574725645594299, 0.466784298885614, 0.124652200378478, 0.0449550023768097, 0.880295827286318, 0.840396371204406, 0.589891077019274, 0.628587051760405, 0.927129407413304, 0.871506726602092, 0.0197337963618338, 0.15528280287981, 0.206014845985919, 0.236760282423347, 0.729445445584133, 0.608500693691894, 0.850427652243525, 0.978474971372634, 0.682678060140461, 0.832968404982239, 0.650393565651029, 0.539665199583396, 0.837687800871208, 0.0949421464465559, 0.181684679351747, 0.0942337135784328, 0.176459646085277, 0.00694482238031924, 0.937692643841729, 0.0936772397253662, 0.988167840056121, 0.917646410642192, 0.496371862944216, 0.884497257415205, 0.728389238705859, 0.842217715922743, 0.639978842111304, 0.614427597727627, 0.824428639840335, 0.427271315827966, 0.760011441772804, 0.741694800788537, 0.921169541077688, 0.301399598130956, 0.420318598393351, 0.36050607659854, 0.590251483954489, 0.584847332211211, 0.591400179080665, 0.962289731949568, 0.139079713029787, 0.431521908612922, 0.414519626647234, 0.96357499435544, 0.818060209508985, 0.253075103508309, 0.354063755366951, 0.457627599360421, 0.714665438048542, 0.919901612447575, 0.0419717174954712, 0.566191446268931, 0.93125799507834, 0.0143440205138177, 0.756214090855792, 0.968428704887629, 0.726468581007794, 0.306167999980971, 0.437175365164876, 0.278290601447225, 0.359401497291401, 0.881793409353122, 0.983965870691463, 0.79985952260904, 0.226586836157367, 0.953777303686365, 0.048837311565876, 0.288561214227229, 0.861007405910641, 0.0409009989816695, 0.586511960253119, 0.800510142697021, 0.234199806116521, 0.126394622027874, 0.0898624826222658, 0.464389294851571, 0.123192323138937, 0.69841585191898, 0.542780998861417, 0.545401087263599, 0.479731506900862, 0.189029246335849, 0.48694116435945, 0.187167347176, 0.981893720803782, 0.606449815677479, 0.258524681208655, 0.298104466870427, 0.261435483349487, 0.888864401960745, 0.766246973769739, 0.804246399318799, 0.123226836323738, 0.217195910634473, 0.738780588144436, 0.800056058447808, 0.237370872870088, 0.177804126869887, 0.149838324636221, 0.0958738352637738, 0.726427505491301, 0.313713403651491, 0.794257765403017, 0.312942957272753, 0.119384094141424, 0.253932501655072, 0.224393437383696, 0.951721490360796, 0.418349927756935, 0.70232285419479, 0.234221554128453, 0.908990857424214, 0.677842316217721, 0.265342441853136, 0.35716388723813, 0.663478885078803, 0.329477587481961, 0.254782811272889, 0.234623775118962, 0.262294941814616, 0.365561816608533, 0.606897032586858, 0.52029159781523, 0.534316891338676, 0.0633123745210469, 0.905677001457661, 0.865663065575063, 0.572905979817733, 0.829079055227339, 0.958101549651474, 0.933307289844379, 0.949690613197163, 0.90848740725778, 0.0672799085732549, 0.592302644392475, 0.429550879169255, 0.243291444377974, 0.983582387678325, 0.144046316621825, 0.560910075670108, 0.253531027818099, 0.351252859458327, 0.797424314543605, 0.684631978161633, 0.484538671560585, 0.595638367114589, 0.137190117733553, 0.894782492890954, 0.305808025645092, 0.431774323107675, 0.0729378762189299, 0.860084173968062, 0.846614917041734, 0.421317723812535, 0.941071244422346, 0.419300701702014, 0.179676304571331, 0.0612725031096488, 0.744677887763828, 0.615542025538161, 0.213658744702116, 0.00802185595966876, 0.769608212867752, 0.833084694109857, 0.171082400484011, 0.717418453656137, 0.978118141880259, 0.259480499429628, 0.968883567489684, 0.689027627231553, 0.258312477264553, 0.631944294786081, 0.583204723894596, 0.289706831797957, 0.588788368040696, 0.712301397230476, 0.321394510800019, 0.753084923373535, 0.531274889828637, 0.704060338437557, 0.0942529682070017, 0.283397865248844, 0.657554040662944, 0.817010851111263, 0.259396306239069, 0.956608986016363, 0.138873098883778, 0.724051113706082, 0.626846161670983, 0.369530040072277, 0.355527567211539, 0.641955336090177, 0.586457871599123, 0.0116500109434128, 0.440462030936033, 0.0695539496373385, 0.554115849547088, 0.222688822075725, 0.844069861341268, 0.906813497887924, 0.638549109920859, 0.768679086817428, 0.435203608125448, 0.560206237249076, 0.736460268264636, 0.473506052512676, 0.945934076560661, 0.448456290410832, 0.221284100553021, 0.317445846041664, 0.124042432522401, 0.342388121644035, 0.11632184474729, 0.702535291668028, 0.00925177335739136, 0.347899759886786, 0.260927308816463, 0.448016530601308, 0.702026563696563, 0.355093490099534, 0.236998612526804, 0.221526912879199, 0.283012266969308, 0.315913542173803, 0.575107142329216, 0.756448655156419, 0.865320363547653, 0.049584039254114, 0.813937530154362, 0.463255458045751, 0.0675225572194904, 0.294928311835974, 0.0848852836061269, 0.190091922646388, 0.746935805538669, 0.833651668159291, 0.299270482733846, 0.546571857761592, 0.134306727675721, 0.0311620596330613, 0.164337298134342, 0.59007621766068, 0.680974962655455, 0.774231750750914, 0.829184416448697, 0.168772722827271, 0.98631036048755, 0.533519905759022, 0.260849867714569, 0.828794787637889, 0.099428990855813, 0.0637389570474625, 0.512330708326772, 0.422052825335413, 0.193742206552997, 0.292626466834918, 0.956220112740993, 0.34930677828379, 0.729049426969141, 0.0276869221124798, 0.401261040475219, 0.638716481393203, 0.596510092495009, 0.591950813774019, 0.719956012209877, 0.758858789457008, 0.0246499450877309, 0.298026784090325, 0.301684509497136, 0.54396872385405, 0.358275031438097, 0.917713976697996, 0.182759156217799, 0.284635856980458, 0.499653368256986, 0.36610939190723, 0.0336572218220681, 0.357285594101995, 0.508448035921901, 0.975088489009067, 0.949916020734236, 0.268106275936589, 0.201804186450318, 0.0284620041493326, 0.97411102685146, 0.295009282184765, 0.0634400234557688, 0.424779338063672, 0.00982740102335811, 0.642554375575855, 0.044959383783862, 0.334041456226259, 0.225525414338335, 0.803286969661713, 0.289794091368094, 0.86934817279689, 0.211567714577541, 0.626558097777888, 0.756669193506241, 0.458922798978165, 0.507967006182298, 0.575722164241597, 0.346772161545232, 0.453396356431767, 0.314341284101829, 0.925803250167519, 0.751140084816143, 0.635851272847503, 0.204985519405454, 0.561666601803154, 0.907210042467341, 0.233278402825817, 0.0844209499191493, 0.206430920399725, 0.184479473624378, 0.352749343728647, 0.032336913747713, 0.104114536195993, 0.449819668196142, 0.172609811881557, 0.589334780583158, 0.0621791176963598, 0.11144298221916, 0.991083640838042, 0.73303681705147, 0.109337414614856, 0.214286357862875, 0.0296175859402865, 0.612292881123722, 0.883614952675998, 0.550966320559382, 0.885762329446152, 0.899008688982576, 0.881736851064488, 0.721829439746216, 0.905008280416951, 0.486117689171806, 0.992978073423728, 0.614267810480669, 0.202391469618306, 0.145829328801483, 0.769359986064956, 0.207533391192555, 0.998546282760799, 0.256635935977101, 0.582679310813546, 0.232271223096177, 0.193877997575328, 0.587823185836896, 0.841342523228377, 0.917976706288755, 0.528663427336141, 0.971990835852921, 0.453681918093935, 0.0616410297807306, 0.550816036295146, 0.670777523657307, 0.484958520857617, 0.615577638149261, 0.690148375695571, 0.940175094874576, 0.0109164065215737, 0.208477455191314, 0.933784739347175, 0.9545813575387, 0.857823513448238, 0.473700639093295, 0.617536254227161, 0.735814641229808, 0.820026914123446, 0.445398871554062, 0.977149189449847, 0.736942684510723, 0.877337311394513, 0.971184203168377, 0.0487887058407068, 0.499809330329299, 0.0837148311547935, 0.0428677827585489, 0.608155848924071, 0.560269837500528, 0.64903089730069, 0.316746459342539, 0.522464264882728, 0.932545140385628, 0.404916713014245, 0.182421293342486, 0.0158037093933672, 0.63306917110458, 0.0974754674825817, 0.65279979724437, 0.13501903694123, 0.863840833771974, 0.432934077223763, 0.680221178103238, 0.537391212536022, 0.0854661106131971, 0.467774659628049, 0.505457759369165, 0.0416403573472053, 0.961671094642952, 0.246536258608103, 0.779640598222613, 0.192669824464247, 0.307285148650408, 0.300952909281477, 0.119089058134705, 0.255910892738029, 0.529689403949305, 0.471346582751721, 0.0949215833097696, 0.833991311257705, 0.572452310007066, 0.0607809389475733, 0.76708904071711, 0.460988988634199, 0.824892183532938, 0.491261418443173, 0.794630954740569, 0.521393080009148, 0.405959653435275, 0.425672769546509, 0.466340172803029, 0.0477387954015285, 0.841788074467331, 0.0317370090633631, 0.879531370243058, 0.0728189554065466, 0.363896788563579, 0.785357154672965, 0.536742906551808, 0.15392466657795, 0.999182515544817, 0.283701779786497, 0.946568689309061, 0.521678041899577, 0.728467038599774, 0.643926066579297, 0.264588043093681, 0.144736886257306, 0.447818540036678, 0.657675775466487, 0.362701824633405, 0.685291432542726, 0.791373910149559, 0.254191041458398, 0.677024442469701, 0.280936094699427, 0.944436252117157, 0.529524773359299, 0.871875873068348, 0.567290857201442, 0.273553194245324, 0.646395086310804, 0.768138346960768, 0.136040005134419, 0.876697512576357, 0.751358800567687, 0.823564473539591, 0.915493807755411, 0.452487520407885, 0.173710107570514, 0.690018117427826, 0.222532809944823, 0.27880319673568, 0.60342507250607, 0.24449918908067, 0.305873323697597, 0.348306666128337, 0.652142733102664, 0.523503744974732, 0.883068537106737, 0.438848403748125, 0.704942867858335, 0.42374153714627, 0.361263899831101, 0.782475450774655, 0.0893536573275924, 0.674523376626894, 0.000112521694973111, 0.887801236705855, 0.60669005732052, 0.970274436054751, 0.182184500619769, 0.11687676422298, 0.194491347996518, 0.600215372629464, 0.992187978001311, 0.867512119235471, 0.429107086965814, 0.743296783650294, 0.802537644514814, 0.462384026497602, 0.142068665008992, 0.984692916972563, 0.760837475536391, 0.204054794274271, 0.59354088944383, 0.966619481565431, 0.0153244149405509, 0.514670911710709, 0.00617827591486275, 0.210747115081176, 0.805379302473739, 0.061665773158893, 0.436999707715586, 0.294718180783093, 0.989551605423912, 0.844474989920855, 0.949651621980593, 0.888196329120547, 0.973027474246919, 0.348079404560849, 0.93379968451336, 0.0307127858977765, 0.257357985246927, 0.707791863707826, 0.483771772822365, 0.655811813427135, 0.530851139919832, 0.0341148225124925, 0.758789101615548, 0.692623148206621, 0.016579233109951, 0.638492221012712, 0.0432726053986698, 0.85962235275656, 0.65136498818174, 0.837005503010005, 0.511500833090395, 0.677765677450225, 0.348060930846259, 0.303239784669131, 0.542775560868904, 0.0127447447739542, 0.781948197167367, 0.334755576681346, 0.980548907304183, 0.94638951565139, 0.136321356752887, 0.586187727982178, 0.563224280020222, 0.597973659867421, 0.530403799144551, 0.252356282668188, 0.964636557735503, 0.525483088335022, 0.946944519178942, 0.24005996552296, 0.869207825744525, 0.984298530500382, 0.980343155330047, 0.372285850346088, 0.351792110828683, 0.235942651983351, 0.742721501970664, 0.0267928552348167, 0.000747995683923364, 0.549387181410566, 0.0877623807173222, 0.548563184449449, 0.582561189774424, 0.526094494387507, 0.546299779321998, 0.967267647851259, 0.344175490085036, 0.626235025003552, 0.166493843076751, 0.715032186359167, 0.0671800337731838, 0.471248008776456, 0.294986988883466, 0.228973179589957, 0.0565729553345591, 0.890917678596452, 0.726820243522525, 0.307434356538579, 0.132658019196242, 0.676419466501102, 0.787763609085232, 0.471743481000885, 0.426011844538152, 0.172342007281259, 0.223707606783137, 0.710847873706371, 0.659568553324789, 0.657378376461565, 0.337599808583036, 0.365727669559419, 0.844621472526342, 0.873809614917263, 0.441546929534525, 0.63304594787769, 0.514305838616565, 0.480883284471929, 0.836603254312649, 0.726493528811261, 0.974542818032205, 0.700199445709586, 0.965160938678309, 0.009637062670663, 0.599894058424979, 0.498204624745995, 0.838384450180456, 0.661087863845751, 0.31311055412516, 0.530575058422983, 0.883496629539877, 0.039552987087518, 0.909705459838733, 0.0228528121951967, 0.973587927175686, 0.0831576811615378, 0.95337966713123, 0.0138997866306454, 0.157210943987593, 0.201893864898011, 0.617154400795698, 0.599345739465207, 0.879246572963893, 0.287317714188248, 0.297872808296233, 0.223959409631789, 0.0701916732359678, 0.172675852663815, 0.416320541873574, 0.909860086627305, 0.32064834726043, 0.19350843434222, 0.0885399274993688, 0.828017135616392, 0.0118251293897629, 0.731281145242974, 0.33697111485526, 0.792243764968589, 0.338670048862696, 0.0380847658962011, 0.779459552606568, 0.272036310983822, 0.0252086429391056, 0.764905496034771, 0.0114341601729393, 0.203519689617679, 0.243016401305795, 0.348268914036453, 0.21938779274933, 0.827169251861051, 0.0517186210490763, 0.885197531199083, 0.718770773615688, 0.914677827386186, 0.0816175458021462, 0.912125644972548, 0.150650619762018, 0.953175035072491, 0.206462528090924, 0.73228175775148, 0.0295779455918819, 0.490522786276415, 0.629802236100659, 0.148538162466139, 0.556460127001628, 0.470066256122664, 0.465587350074202, 0.781614177394658, 0.15811773086898, 0.713390670251101, 0.455883125541732, 0.335618294775486, 0.493253686698154, 0.394107156898826, 0.897874503396451, 0.732207027031109, 0.139543692348525, 0.192007077625021, 0.850927577586845, 0.190567325800657, 0.163780124625191, 0.399394648848101, 0.822744844481349, 0.93784850789234, 0.356110208667815, 0.287592037813738, 0.534963868791237, 0.539524567313492, 0.906128829112276, 0.200743496185169, 0.443830723175779, 0.0621356491465122, 0.419203293509781, 0.0358716640621424, 0.15943625732325, 0.381225140532479, 0.266330732498318, 0.0746970537584275, 0.134599615354091, 0.733151084277779, 0.717752740252763, 0.506466323276982, 0.468934744596481, 0.699450999498367, 0.567634453531355, 0.435076334048063, 0.359113342827186, 0.645311094122007, 0.869958489201963, 0.341129907639697, 0.856825932161883, 0.806160556152463, 0.0622365679591894, 0.145404225448146, 0.0920309845823795, 0.928949529770762, 0.374735079007223, 0.634937633993104, 0.328409814508632, 0.0929342294111848, 0.483777579152957, 0.637775024631992, 0.748740399489179, 0.609062801115215, 0.348996306071058, 0.248236172832549, 0.0703158359974623, 0.0978052779100835, 0.703789631370455, 0.649434860562906, 0.535098241409287, 0.0854632570408285, 0.0123003760818392, 0.219909422798082, 0.632196955149993, 0.162394913379103, 0.964770303340629, 0.744367084465921, 0.0344960496295244, 0.929102738387883, 0.811612512916327, 0.236931428313255, 0.897038796218112, 0.0802146133501083, 0.56991378380917, 0.204128979705274, 0.152536438778043, 0.888025194872171, 0.25945828598924, 0.0236999921035022, 0.632605588529259, 0.848563662264496, 0.229850866133347, 0.146350236609578, 0.832226770464331, 0.356928643072024, 0.308354628970847, 0.514780693221837, 0.0467977579683065, 0.141928493743762, 0.965631127357483, 0.818664524471387, 0.239806927740574, 0.312129940837622, 0.427358304616064, 0.000490188598632812, 0.105297592002898, 0.865464511094615, 0.616597461514175, 0.267067705513909, 0.501462459564209, 0.851929078111425, 0.294398497324437, 0.33502703323029, 0.709991284646094, 0.24412423861213, 0.703214361099526, 0.0930609509814531, 0.602827673777938, 0.293184748617932, 0.684381478466094, 0.370932373451069, 0.874512086389586, 0.125178206013516, 0.0836786876898259), actual = c(0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.0520731636788696, 0.774613732239231, 0.258852173807099, 0.668674060609192, 0.908133858814836, 0.0303622097708285, 0.309606282273307, 0.764418041100726, 0.930290490621701, 0.0873995099682361, 0.412269626976922, 0.912364655872807, 0.550389781827107, 0.224116111639887, 0.0122739649377763, 0.45910069369711, 0.512025133706629, 0.496479156659916, 0.0341621804982424, 0.0430422127246857, 0.252064065309241, 0.19199652178213, 0.917292989091948, 0.114443533588201, 0.332060520537198, 0.984718061052263, 0.581118049332872, 0.730584692908451, 0.547142881667241, 0.48611703584902, 0.896088112145662, 0.793971504783258, 0.877798018045723, 0.605223898543045, 0.483813709346578, 0.0171124082989991, 0.265706979203969, 0.789289504289627, 0.343780806055292, 0.0103193465620279, 0.146955392789096, 0.203216970199719, 0.449102875078097, 0.813661837717518, 0.153582723112777, 0.425598427187651, 0.388598901452497, 0.29852943867445, 0.549862071173266, 0.910981429042295, 0.275586113566533, 0.150194109883159, 0.945444298209623, 0.784539836458862, 0.470415152376518, 0.0746462394017726, 0.127310644136742, 0.969164189882576, 0.700684800278395, 0.365958391688764, 0.587738121859729, 0.46629804559052, 0.963416135171428, 0.304352911189198, 0.165025054244325, 0.285625993041322, 0.6218870007433, 0.96496977400966, 0.66591797536239, 0.633131038397551, 0.31994592002593, 0.89626772259362, 0.624145960202441, 0.905668482650071, 0.342623368836939, 0.932117245858535, 0.249048347352073, 0.696665910305455, 0.592345780460164, 0.133484638528898, 0.441154743544757, 0.142831909237429, 0.822987295920029, 0.688621176639572, 0.575075228698552, 0.846833884948865, 0.134009503992274, 0.844233844894916, 0.522769623668864, 0.837967450963333, 0.780913983471692, 0.78051751316525, 0.43274229997769, 0.780286229448393, 0.571693511446938, 0.534037577686831, 0.722604256588966, 0.682608266826719, 0.489839307963848, 0.459612111793831, 0.411461492301896, 0.589157008798793, 0.882824296364561, 0.543656344525516, 0.751251090317965, 0.357643488794565, 0.872079024557024, 0.737292011035606, 0.468241453403607, 0.728589711012319, 0.909049336332828, 0.667228441918269, 0.14510747930035, 0.886642551282421, 0.020151169737801, 0.512997813755646, 0.586689660092816, 0.701189976418391, 0.263836125610396, 0.313882740912959, 0.324918538099155, 0.0754284844733775, 0.918162704911083, 0.519220564281568, 0.828343258937821, 0.341017007827759, 0.110363751417026, 0.172799789579585, 0.385152826551348, 0.214042310370132, 0.73906838381663, 0.654974992386997, 0.556307910010219, 0.3241114274133, 0.311996927717701, 0.53897351725027, 0.933666392229497, 0.826948906295002, 0.15071654622443, 0.890821647597477, 0.715441960375756, 0.24448095750995, 0.948457397520542, 0.927036018809304, 0.362994702998549, 0.208245773799717, 0.534620012156665, 0.729874662123621, 0.0369082242250443, 0.291732143843547, 0.57477134722285, 0.183802840765566, 0.499722854699939, 0.0646188715472817, 0.0333721088245511, 0.723608169704676, 0.779767524916679, 0.104504431597888, 0.960998520255089, 0.729398540453985, 0.248620525933802, 0.414688198128715, 0.377023108769208, 0.695170937571675, 0.669611131306738, 0.913127555744722, 0.943496454274282, 0.798161627026275, 0.469367585610598, 0.0232355995103717, 0.292213334934786, 0.93468480091542, 0.156413706019521, 0.496053622104228, 0.235336703015491, 0.00540849776007235, 0.200498502701521, 0.79012665594928, 0.739553263643757, 0.450975421350449, 0.527662663953379, 0.429168775910512, 0.934766662539914, 0.568223343230784, 0.289985067443922, 0.594012815039605, 0.966535183368251, 0.163218213710934, 0.250538215506822, 0.252627388108522, 0.0323390197008848, 0.36726885358803, 0.0421585601288825, 0.294793611392379, 0.49235420464538, 0.422241527587175, 0.159639877267182, 0.75300105381757, 0.143429556628689, 0.79119048640132, 0.436866228934377, 0.653853487456217, 0.366176036186516, 0.633993176743388, 0.826395421754569, 0.304093239828944, 0.0403284248895943, 0.411930947098881, 0.438379396684468, 0.193919947603717, 0.121600991580635, 0.575468043331057, 0.121120526688173, 0.387534967856482, 0.647499285638332, 0.651531362906098, 0.363531351555139, 0.98766168858856, 0.654083904111758, 0.684375817654654, 0.488203227519989, 0.634609933244064, 0.230872554006055, 0.691771952668205, 0.553007858572528, 0.534593390766531, 0.231436502654105, 0.749072067206725, 0.75149061717093, 0.39835072751157, 0.329078370938078, 0.515377060975879, 0.189245279645547, 0.372483156155795, 0.0211261278018355, 0.764728114241734, 0.528608000837266, 0.963334185769781, 0.553007559850812, 0.357080666115507, 0.92976806499064, 0.913503486197442, 0.920530670788139, 0.15453161043115, 0.795539531856775, 0.0243018337059766, 0.859098721761256, 0.3707734211348, 0.0902577897068113, 0.356661133700982, 0.27196105546318, 0.499755061930045, 0.50272056623362, 0.378863421967253, 0.0228086593560874, 0.267503486946225, 0.815670750103891, 0.366642446490005, 0.685482931323349, 0.513593556592241, 0.265526456525549, 0.0732827608007938, 0.896984503371641, 0.932955064810812, 0.175197620177642, 0.677891377126798, 0.150396064389497, 0.293244210071862, 0.977950193220749, 0.123324340442196, 0.613093447871506, 0.81296289130114, 0.466896815225482, 0.147149230353534, 0.479463391937315, 0.340321261202917, 0.586104131303728, 0.77127645816654, 0.755313072353601, 0.284244898008183, 0.477129822829738, 0.509571342496201, 0.478221810190007, 0.512784488033503, 0.16012162505649, 0.205090924864635, 0.687404795782641, 0.22471377113834, 0.427454597549513, 0.904099953360856, 0.0869340216740966, 0.724714362295344, 0.527887812117115, 0.108787407167256, 0.80817271117121, 0.83229545969516, 0.932236544089392, 0.194412948796526, 0.0235925333108753, 0.65218848711811, 0.971323052654043, 0.802586935693398, 0.0576539309695363, 0.185155659681186, 0.351698357611895, 0.785247757332399, 0.261369514744729, 0.574202029965818, 0.445797954453155, 0.384883964201435, 0.63981849886477, 0.675128175644204, 0.847657172707841, 0.41272360063158, 0.554656078340486, 0.242118299473077, 0.00338041572831571, 0.414610343053937, 0.285827762912959, 0.751324224751443, 0.417267105542123, 0.75232992367819, 0.159762215102091, 0.655275470810011, 0.669023503083736, 0.628972273087129, 0.83792720339261, 0.587412498891354, 0.372358703287318, 0.159179847687483, 0.137632185593247, 0.161699593067169, 0.537801766768098, 0.0348540374543518, 0.586356045445427, 0.207804162288085, 0.0857436193618923, 0.479892184026539, 0.081378776114434, 0.925341820111498, 0.329558691941202, 0.0420792719814926, 0.44130876311101, 0.714997890638188, 0.727746590739116, 0.34599073487334, 0.58094075601548, 0.490133555838838, 0.924306998029351, 0.591601417399943, 0.813564697513357, 0.534750433173031, 0.696359020890668, 0.0410762180108577, 0.278920858399943, 0.0585397153627127, 0.88733583712019, 0.354270439594984, 0.390622636536136, 0.549228445161134, 0.715135222533718, 0.448477737838402, 0.963533007306978, 0.308954526903108, 0.589063552673906, 0.924362349323928, 0.899216324556619, 0.33297995198518, 0.278478007763624, 0.187469360185787, 0.930619245627895, 0.548791645560414, 0.392541626933962, 0.913243510993198, 0.764338227687404, 0.293614980997518, 0.620325637049973, 0.933100785594434, 0.788851204095408, 0.0719400860834867, 0.623482759809121, 0.706291066715494, 0.475840683560818, 0.507741595385596, 0.166657358873636, 0.644170769955963, 0.79032274428755, 0.91281860996969, 0.248345782514662, 0.0505458682309836, 0.515160947339609, 0.729818791151047, 0.678644808474928, 0.950090513098985, 0.889740463811904, 0.411904128268361, 0.756207626778632, 0.396074193064123, 0.908521325327456, 0.0571889965794981, 0.200492408825085, 0.852550379466265, 0.251406393013895, 0.995968880830333, 0.595422715181485, 0.600525428541005, 0.437948522623628, 0.884460556553677, 0.317434829194099, 0.180008220020682, 0.0751079765614122, 0.00181807111948729, 0.923627582145855, 0.538978626718745, 0.629579711705446, 0.346991059836, 0.480119379237294, 0.744873341405764, 0.382823082618415, 0.254195224260911, 0.652641949942335, 0.46417623735033, 0.679594748653471, 0.086660306667909, 0.400059866718948, 0.238196751102805, 0.388839178718626, 0.910822091624141, 0.07121567777358, 0.318009350681677, 0.296729200286791, 0.607536964584142, 0.115469966316596, 0.450449516065419, 0.635854383930564, 0.467365138232708, 0.279835696797818, 0.293276007520035, 0.173553233034909, 0.241427805507556, 0.0444293073378503, 0.552078674314544, 0.494551982730627, 0.443406093632802, 0.341934111667797, 0.0843645115382969, 0.615587410284206, 0.16181420465, 0.374578086426482, 0.288259514374658, 0.384903795085847, 0.114690142218024, 0.980264206416905, 0.467041322030127, 0.679267236264423, 0.997256741160527, 0.354588843882084, 0.722464395686984, 0.936812193132937, 0.4991514547728, 0.277517667040229, 0.346314071211964, 0.751463281922042, 0.723706628661603, 0.46675501158461, 0.744737227680162, 0.724140227073804, 0.298433571588248, 0.594482944346964, 0.852972219698131, 0.224082744680345, 0.287514344323426, 0.778841271065176, 0.484431995544583, 0.677764064166695, 0.878084788331762, 0.516335308086127, 0.778323269449174, 0.869936334434897, 0.807350638555363, 0.912381667178124, 0.884656528476626, 0.138576369965449, 0.103298209607601, 0.731389736291021, 0.155552218900993, 0.143287983955815, 0.867755665211007, 0.652787521714345, 0.877214646665379, 0.815150327049196, 0.993751103058457, 0.404779080068693, 0.751224277541041, 0.327196437865496, 0.487129029119387, 0.563529718201607, 0.839586000656709, 0.551626898115501, 0.000645978143438697, 0.649744236608967, 0.493422711733729, 0.419412495801225, 0.938346593873575, 0.992415082640946, 0.468560197390616, 0.741151979658753, 0.565152045572177, 0.102771794423461, 0.213150778552517, 0.528448048513383, 0.303294952493161, 0.0825673937797546, 0.213532725349069, 0.149465812370181, 0.18181836581789, 0.200829740148038, 0.19089766475372, 0.0784290188457817, 0.110674172407016, 0.111861585872248, 0.451929129194468, 0.961162757594138, 0.212767392862588, 0.282006888650358, 0.000713243382051587, 0.228827025275677, 0.350806010654196, 0.280830432195216, 0.145122393267229, 0.561910884687677, 0.998233583755791, 0.791327884886414, 0.658465815708041, 0.783241000026464, 0.688373975688592, 0.655750266741961, 0.745319510111585, 0.733388565247878, 0.915885997004807, 0.368183989776298, 0.749754713149741, 0.729982287622988, 0.169785106554627, 0.0231048194691539, 0.30501218046993, 0.822382186772302, 0.841318373568356, 0.127847017720342, 0.553130525164306, 0.492708407109603, 0.549485540250316, 0.287585937883705, 0.328515256987885, 0.325437674531713, 0.586685492889956, 0.825436293613166, 0.664547959342599, 0.457659694831818, 0.181102424161509, 0.230291555868462, 0.0809006518684328, 0.917977734468877, 0.649255078984424, 0.733229232253507, 0.877582119079307, 0.330637098290026, 0.633789606625214, 0.189456143649295, 0.362449169391766, 0.827070976374671, 0.224283179501072, 0.240465571638197, 0.151543905027211, 0.211056355154142, 0.986741245025769, 0.593440218130127, 0.529780013021082, 0.397040571086109, 0.727558166487142, 0.471307694446295, 0.873699301853776, 0.863123409217224, 0.315120037877932, 0.166362969204783, 0.633072290103883, 0.48375659622252, 0.806440571090207, 0.566647425526753, 0.0551681600045413, 0.303469164296985, 0.0738988663069904, 0.0265156906098127, 0.787252781214193, 0.848340106429532, 0.34884696896188, 0.756616736529395, 0.134739023400471, 0.550302082672715, 0.27928410237655, 0.597079882398248, 0.717212948249653, 0.589025760069489, 0.662488030968234, 0.740332254441455, 0.0313967005349696, 0.153827011818066, 0.868040183791891, 0.887596332002431, 0.608579811872914, 0.300802100682631, 0.062494623940438, 0.387477067299187, 0.730510594323277, 0.756331202806905, 0.626893671927974, 0.500106660882011, 0.336771237198263, 0.267377808690071, 0.106963122496381, 0.555819413857535, 0.420170158846304, 0.236596956383437, 0.374771837843582, 0.18691006815061, 0.375490531791002, 0.631310187047347, 0.273901585955173, 0.716515736887231, 0.442612352780998, 0.229803247144446, 0.993344114394858, 0.427057720255107, 0.654431633884087, 0.776920821052045, 0.955659894039854, 0.171996597200632, 0.689249514369294, 0.373948925640434, 0.964123240439221, 0.127250783378258, 0.73267278354615, 0.222197760362178, 0.637510257773101, 0.247098307823762, 0.481514737010002, 0.50626910221763, 0.399223676882684, 0.290701500140131, 0.300904008792713, 0.0270867755170912, 0.765882266452536, 0.334867853205651, 0.353819349315017, 0.841216516215354, 0.378365239594132, 0.0414862763136625, 0.573006273247302, 0.114420137600973, 0.882077243179083, 0.738892063964158, 0.836477028671652, 0.542928119655699, 0.606806413736194, 0.418136277003214, 0.796102334279567, 0.462812717771158, 0.0760270797181875, 0.348838637582958, 0.00541405053809285, 0.901417322922498, 0.573969596298411, 0.85195947973989, 0.35248862975277, 0.732928075594828, 0.426521454006433, 0.0255326493643224, 0.656545143807307, 0.199977841461077, 0.945818507112563, 0.821832515532151, 0.295665504876524, 0.486066818702966, 0.167092491174117, 0.637542854761705, 0.144738490460441, 0.967914446955547, 0.875061715021729, 0.475101584102958, 0.791542893508449, 0.840459985425696, 0.406097492435947, 0.0669793426059186, 0.577301366487518, 0.738202937180176, 0.764850097009912, 0.923943430185318, 0.0692742150276899, 0.230449448805302, 0.47026333026588, 0.181243841070682, 0.590116220992059, 0.455338043160737, 0.344628771999851, 0.991526113590226, 0.705659211147577, 0.0849950916599482, 0.421482791891322, 0.0598801339510828, 0.360084047075361, 0.031177731230855, 0.154170991154388, 0.31659331661649, 0.168597855372354, 0.360179882962257, 0.64393571880646, 0.391968272393569, 0.3798320107162, 0.25351573410444, 0.589397182222456, 0.741025461815298, 0.825376400491223, 0.912536759162322, 0.0365944355726242, 0.875369055662304, 0.781699903542176, 0.679435824276879, 0.0229764657560736, 0.0404621849302202, 0.381388830719516, 0.128628758015111, 0.391432389849797, 0.748120461357757, 0.46937333792448, 0.3537255639676, 0.480735369259492, 0.0258808548096567, 0.21419446519576, 0.265744176926091, 0.667456038994715, 0.499529263703153, 0.768456469057128, 0.273956290213391, 0.113164026290178, 0.101339862681925, 0.25405095377937, 0.899200660409406, 0.461164605803788, 0.335392975481227, 0.925080104498193, 0.179835924878716, 0.192720019025728, 0.122751188464463, 0.241678084945306, 0.493456397904083, 0.877051230985671, 0.0100290887057781, 0.267211942700669, 0.261647200444713, 0.539938393747434, 0.760900612454861, 0.00584975304082036, 0.174176550237462, 0.703859533881769, 0.969153338111937, 0.121280951891094, 0.908893608022481, 0.851832680171356, 0.330955345882103, 0.404296206077561, 0.126244311220944, 0.603422920918092, 0.238671897677705, 0.605378963518888, 0.800191141664982, 0.157642285805196, 0.805049142800272, 0.104171717073768, 0.638137370813638, 0.20448022801429, 0.616842550691217, 0.191376441391185, 0.64694365253672, 0.486639238428324, 0.613202654058114, 0.754171684617177, 0.0623958627693355, 0.944984721951187, 0.491905958857387, 0.964198790257797, 0.318787133321166, 0.951989985769615, 0.846986782271415, 0.58104422967881, 0.573416263796389, 0.280311761656776, 0.804496776778251, 0.97683810419403, 0.218508012127131, 0.33593900827691, 0.201341063948348, 0.912243851227686, 0.993216281523928, 0.0583848007954657, 0.881831207545474, 0.887746432330459, 0.878072732128203, 0.665075884899125, 0.763668816769496, 0.845353460172191, 0.788822608534247, 0.859267238294706, 0.471813977928832, 0.165142690762877, 0.117678553331643, 0.359863726422191, 0.903067487524822, 0.49438398052007, 0.583006331929937, 0.405661103315651, 0.169185638194904, 0.332529379520565, 0.55029074056074, 0.377451218198985, 0.161456543253735, 0.440884046955034, 0.543634777422994, 0.154546377249062, 0.420334339840338, 0.381072961492464, 0.595273935701698, 0.452346077654511, 0.692628055112436, 0.615355719113722, 0.710576735669747, 0.397039233474061, 0.817581585608423, 0.766817959491163, 0.839039304992184, 0.749525868101045, 0.954908726736903, 0.412944386014715, 0.0474690082482994, 0.197336332406849, 0.34643319924362, 0.0321319501381367, 0.542156149633229, 0.980129071045667, 0.0852840628940612, 0.655076311901212, 0.842796789482236, 0.911414043279365, 0.510766853811219, 0.435210619354621, 0.155812873039395, 0.97916665696539, 0.50303738634102, 0.461807396030053, 0.0611491589806974, 0.0323574235662818, 0.165522624971345, 0.59029716392979, 0.79889539629221, 0.00792537722736597, 0.0687964756507427, 0.36528892419301, 0.00757184438407421, 0.131079454440624, 0.627931703114882, 0.82505848701112, 0.184886880684644, 0.786685210186988, 0.694099347339943, 0.321481806458905, 0.474474772578105, 0.118527684127912, 0.306591984815896, 0.835364639526233, 0.317590520251542, 0.768109483178705, 0.426055833464488, 0.3169572558254, 0.463728214381263, 0.303831205470487, 0.222488559316844, 0.474111081101, 0.846273462055251, 0.971441246103495, 0.312269270652905, 0.934609145158902, 0.286758247995749, 0.794419148005545, 0.664865067927167, 0.567333427025005, 0.644189736340195, 0.940029081655666, 0.534357760567218, 0.421196865383536, 0.471712002996355, 0.248039880767465, 0.569219185272232, 0.933501966064796, 0.710331379668787, 0.360110366949812, 0.270691997371614, 0.420778995845467, 0.628308325307444, 0.260421532671899, 0.293827142100781, 0.446894804015756, 0.76531415223144, 0.618992869975045, 0.207099444465712, 0.566205074777827, 0.948791271541268, 0.0237185980658978, 0.406560161383823, 0.505571344867349, 0.0928501421585679, 0.132736499188468, 0.428519446868449, 0.854538434185088, 0.962436361704022, 0.980683425907046, 0.394687861902639, 0.463550481945276, 0.118064328329638, 0.373935357900336, 0.188287522876635, 0.542521242517978, 0.562641217606142, 0.167902249610052, 0.184183985227719, 0.337128327228129, 0.667113667353988, 0.287142521003261, 0.0676666372455657, 0.200406070798635, 0.488608815474436, 0.757553570903838, 0.725503021618351, 0.0512735741212964, 0.0827291284222156, 0.137038927758113, 0.0545389673206955, 0.572431232081726, 0.668486122041941, 0.840838954085484, 0.135416397824883, 0.669979950413108, 0.581657952396199, 0.0614465184044093, 0.140025606611744, 0.916068307356909, 0.096969174221158, 0.90360767277889, 0.804411001037806, 0.713955236133188, 0.453409047331661, 0.524927124613896, 0.0646943729370832, 0.723366703139618, 0.310729596763849, 0.417709069093689, 0.100155759137124, 0.652496714377776, 0.518465495901182, 0.150168888270855, 0.320734736509621, 0.692016380606219, 0.582818689057603, 0.369105599820614, 0.272024234058335, 0.578442238271236, 0.721174941863865, 0.206502778222784, 0.211866995785385, 0.130964327603579, 0.882952202577144, 0.477007330395281, 0.420341338962317, 0.603238601004705, 0.71826147264801, 0.409020429244265, 0.891785625834018, 0.679908186197281, 0.266094866674393, 0.991199646377936, 0.0258442517369986, 0.938693682197481, 0.593242105096579, 0.663949334062636, 0.243597246939316, 0.655466314638034, 0.117862754035741, 0.121421272633597, 0.150108674541116, 0.415011137491092, 0.301323959138244, 0.938421976054087), actual = c(1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"