Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.9345912        0.9345912 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4464408        0.4464408        0.4772760 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.0266144741326571, 0.976012354716659, 0.531143863452598, 0.604452739236876, 0.677808961365372, 0.862152957823128, 0.257671711267903, 0.88375702756457, 0.85451293271035, 0.768334209918976, 0.72406596224755, 0.381822980474681, 0.676730630220845, 0.553344367304817, 0.965700709726661, 0.0959466062486172, 0.330810207407922, 0.302125425776467, 0.203843800583854, 0.0714619557838887, 0.11754833930172, 0.27024833788164, 0.871316527482122, 0.724882740527391, 0.463036257773638, 0.374815305229276, 0.395653046900406, 0.991216047666967, 0.299947893479839, 0.435310018714517, 0.799113054992631, 0.412176043028012, 0.406597419641912, 0.346654839813709, 0.804821825586259, 0.0463298980612308, 0.671337204985321, 0.293948137899861, 0.529136710800231, 0.401329226326197, 0.449070093221962, 0.486120223999023, 0.692934676073492, 0.79783801548183, 0.881467814091593, 0.871135394321755, 0.541255845921114, 0.330165971303359, 0.0346264450345188, 0.259661314543337, 0.422913814894855, 0.211600732989609, 0.175143678439781, 0.676828620489687, 0.300176249118522, 0.0881063293199986, 0.437023345148191, 0.174558842089027, 0.684285131748766, 0.856723330682144, 0.700830599525943, 0.377910011913627, 0.037739023566246, 0.452244731830433, 0.162747926078737, 0.862880297238007, 0.842946096556261, 0.930693649686873, 0.806692684767768, 0.366163271246478, 0.577747527044266, 0.140986357117072, 0.0783014036715031, 0.0455191908404231, 0.84387260582298, 0.349998750258237, 0.645981139503419, 0.902494049398229, 0.795557511039078, 0.494041089667007, 0.137886255513877, 0.947476235684007, 0.739991270238534, 0.1938034996856, 0.245218451600522, 0.390962430043146, 0.479218815919012, 0.514135029166937, 0.635755914496258, 0.87166532734409, 0.00219665444456041, 0.813474904280156, 0.198714509606361, 0.826556476997212, 0.697965536499396, 0.800551530905068, 0.347834465792403, 0.79620979516767, 0.102270297938958, 0.223654225934297, 0.498011776944622, 0.973105197539553, 0.569480370962992, 0.756541634909809, 0.811311125522479, 0.713635259773582, 0.872390031581745, 0.0142166106961668, 0.841990048298612, 0.362957207486033, 0.943274112185463, 0.739148123888299, 0.136511599877849, 0.439062926918268, 0.61962164985016, 0.643444153713062, 0.992755830986425, 0.125487395795062, 0.652854437241331, 0.140759591944516, 0.0219819804187864, 0.114633091492578, 0.909486145246774, 0.695435780100524, 0.212862980784848, 0.328542906558141, 0.106101568089798, 0.546826141886413, 0.244686202611774, 0.625960804289207, 0.906388859730214, 0.221815603785217, 0.0537390736863017, 0.896891844458878, 0.28936144313775, 0.0195818096399307, 0.531795591581613, 0.499739941209555, 0.853861588751897, 0.73398559098132, 0.085754272993654, 0.932007259922102, 0.753310548141599, 0.244598294142634, 0.0466160462237895, 0.216909733135253, 0.109415617771447, 0.229057303862646, 0.499228346627206, 0.46653897408396, 0.417176271555945, 0.646612586453557, 0.713209603214636, 0.48240470699966, 0.423746506450698, 0.289065793156624, 0.522053291089833, 0.545349200488999, 0.39079915988259, 0.332084030844271, 0.9452832902316, 0.467228743014857, 0.660323990974575, 0.445091055938974, 0.729689044877887, 0.85547752212733, 0.023148778360337, 0.767936740070581, 0.455496381502599, 0.00135227479040623, 0.797201926354319, 0.867111926199868, 0.893076402833685, 0.0886559977661818, 0.613016563933343, 0.967309650732204, 0.265736697474495, 0.0849628245923668, 0.807904018787667, 0.757759032538161, 0.550836404785514, 0.93795476667583, 0.742270938819274, 0.0653078057803214, 0.5459460017737, 0.258696493227035, 0.892360621597618, 0.642150909407064, 0.0702134210150689, 0.0173911894671619, 0.870639791479334, 0.318315399345011, 0.0398178820032626, 0.353937258711085, 0.409843715373427, 0.153454978018999, 0.11193290958181, 0.153693131403998, 0.978562817210332, 0.925283043179661, 0.0355203212238848, 0.129484310280532, 0.0994096859358251, 0.812939551658928, 0.884347456973046, 0.888927592895925, 0.725082129240036, 0.779250199440867, 0.0832875394262373, 0.200207864865661, 0.641822790494189, 0.957673338940367, 0.199328197166324, 0.395348468096927, 0.696908143348992, 0.703468554420397, 0.921347474912181, 0.138663633493707, 0.201466890750453, 0.838799477322027, 0.939590560737997, 0.572376758791506, 0.603519481839612, 0.419342855457217, 0.91090859961696, 0.0983682777732611, 0.127044369000942, 0.312540758633986, 0.756763619603589, 0.841941628139466, 0.746308653848246, 0.140112104592845, 0.719377134460956, 0.522090022452176, 0.878317527472973, 0.947595380945131, 0.256638880586252, 0.57869535824284, 0.759319597156718, 0.21572928619571, 0.328370813978836, 0.662592440610752, 0.590632790699601, 0.15572749148123, 0.808114944491535, 0.859761312138289, 0.376872805645689, 0.0602789607364684, 0.629339321749285, 0.195027548121288, 0.021700392011553, 0.986539319623262, 0.729962554527447, 0.466700945282355, 0.688698462210596, 0.742406792240217, 0.126580094918609, 0.768174188910052, 0.584716280223802, 0.378876379691064, 0.0399202015250921, 0.239935158519074, 0.566946375183761, 0.919925446622074, 0.123433332424611, 0.80021477676928, 0.77209343877621, 0.756859192624688, 0.474439818179235, 0.0218907447997481, 0.144534364808351, 0.648382884217426, 0.464955588802695, 0.733264409471303, 0.977262821048498, 0.773099420359358, 0.832115575671196, 0.517862912733108, 0.478692844277248, 0.353616952663288, 0.788481337949634, 0.465614333748817, 0.488284293795004, 0.17906805826351, 0.107913325773552, 0.568113397341222, 0.0664339626673609, 0.207554501015693, 0.170575397321954, 0.843269313452765, 0.430160137126222, 0.981271962169558, 0.135418695397675, 0.646616216050461, 0.144400204764679, 0.988984255120158, 0.0417444026097655, 0.694056774489582, 0.471752325538546, 0.285978444153443, 0.854650021297857, 0.0289948997087777, 0.126294559100643, 0.72838391829282, 0.168160150526091, 0.0308152807410806, 0.613743732217699, 0.243498344672844, 0.226634747348726, 0.437444997252896, 0.0794141222722828, 0.443375137168914, 0.865711888531223, 0.794509466970339, 0.840368358651176, 0.478842957410961, 0.755679037421942, 0.127670233370736, 0.730211622547358, 0.749347888631746, 0.911159495590255, 0.31000035488978, 0.957301826681942, 0.551589461276308, 0.827676850603893, 0.365461868932471, 0.237627319525927, 0.798846272286028, 0.438696320168674, 0.582522717304528, 0.306513690156862, 0.152506636222824, 0.292696323944256, 0.408349416684359, 0.102560279192403, 0.107868568040431, 0.0861391285434365, 0.516891694162041, 0.386875322787091, 0.956103680189699, 0.895999343832955, 0.325878137024119, 0.181647466495633, 0.989323669346049, 0.285459297709167, 0.143890187144279, 0.0820160859730095, 0.689653607085347, 0.385909304721281, 0.676729237893596, 0.778621584875509, 0.0497263150755316, 0.96362419705838, 0.197035225806758, 0.454940338851884, 0.00593451294116676, 0.434143932303414, 0.48358619469218, 0.295166686875746, 0.322049258975312, 0.366673403885216, 0.0779055268503726, 0.191248430404812, 0.727316379081458, 0.415696850046515, 0.165189311141148, 0.080473953159526, 0.726405259920284, 0.181668671779335, 0.350923096528277, 0.0323787771631032, 0.515971441520378, 0.478243087884039, 0.491491538006812, 0.724003300536424, 0.89600991178304, 0.294824907556176, 0.494306202977896, 0.754609440453351, 0.161747470730916, 0.21753465058282, 0.484625751618296, 0.556624532211572, 0.898984530707821, 0.582003153162077, 0.166807856177911, 0.906736891949549, 0.19996966724284, 0.582805450307205, 0.592257265234366, 0.647090968908742, 0.875586438458413, 0.0278813459444791, 0.109277079347521, 0.50558793079108, 0.0323828035034239, 0.596466196235269, 0.0946118365973234, 0.233070359099656, 0.622939919587225, 0.559510867344216, 0.570606887107715, 0.0105989756993949, 0.814720052294433, 0.765134095447138, 0.189927886938676, 0.961923640919849, 0.712709420826286, 0.950318627757952, 0.766023063566536, 0.135299860034138, 0.822832463541999, 0.17174476548098, 0.14396109059453, 0.556070044869557, 0.366799544775859, 0.923912993865088, 0.934755298774689, 0.0122493919916451, 0.0610552504658699, 0.203709927154705, 0.18499651667662, 0.483356339856982, 0.122290248749778, 0.396857360377908, 0.59668852807954, 0.306134575046599, 0.496105164289474, 0.645814078859985, 0.390643751248717, 0.926730382954702, 0.669487366918474, 0.786726662423462, 0.559035655576736, 0.632140962872654, 0.812766573624685, 0.548642603913322, 0.811037109233439, 0.451936320634559, 0.139819835312665, 0.0777731433045119, 0.532755736494437, 0.60387899982743, 0.785583721706644, 0.17158158053644, 0.278836783953011, 0.887274779146537, 0.978206690633669, 0.63487261813134, 0.235585951944813, 0.656134408665821, 0.305109620559961, 0.583776809973642, 0.906703873770311, 0.202913251239806, 0.706887657754123, 0.656867831945419, 0.609172922791913, 0.631275838473812, 0.455516759306192, 0.124422803521156, 0.745650178520009, 0.239843693561852, 0.501833185087889, 0.0143677908927202, 0.0134170644450933, 0.907455889973789, 0.0636459493543953, 0.336327063152567, 0.384181127185002, 0.267945738974959, 0.749659570399672, 0.91622677189298, 0.338608250021935, 0.793880263576284, 0.245783129474148, 0.434375057229772, 0.240717656444758, 0.0682299437467009, 0.410824705380946, 0.254148582695052, 0.295599339762703, 0.1835777903907, 0.618549679871649, 0.463961165864021, 0.899789978982881, 0.593912166310474, 0.785078766057268, 0.816090570995584, 0.694268602645025, 0.805663416627795, 0.000827444717288017, 0.185735423583537, 0.635313893668354, 0.713977108476683, 0.404590144520625, 0.427325468976051, 0.343021324370056, 0.597659506369382, 0.81809737556614, 0.530029682675377, 0.803890373557806, 0.904524722835049, 0.978281691903248, 0.0805030721239746, 0.729025871958584, 0.361799706704915, 0.802956300787628, 0.44057014817372, 0.100947503233328, 0.0985795420128852, 0.136937707895413, 0.239583821967244, 0.476601520553231, 0.0932749074418098, 0.0843989211134613, 0.0588960980530828, 0.980610548518598, 0.353597369045019, 0.196038876892999, 0.182582643581554, 0.575078014051542, 0.578881359891966, 0.144036999437958, 0.373931192792952, 0.564613641472533, 0.778360008262098, 0.693374528316781, 0.105159782106057, 0.0995843750424683, 0.934972866903991, 0.0583658227697015, 0.109351002844051, 0.807969489134848, 0.974962430773303, 0.0928460475988686, 0.900413051480427, 0.231901590013877, 0.836925823474303, 0.275986257009208, 0.982366069220006, 0.247396844439209, 0.750634007155895, 0.695191801292822, 0.311350790550932, 0.81803595344536, 0.993767145322636, 0.155003177234903, 0.0499714813195169, 0.302723478758708, 0.64170055068098, 0.996996184578165, 0.822890191804618, 0.306440072599798, 0.715654434869066, 0.0127170302439481, 0.943335063755512, 0.682701049605384, 0.632745518349111, 0.619631186593324, 0.789051420753822, 0.85091828648001, 0.126954591833055, 0.3942677276209, 0.635524653596804, 0.974099488463253, 0.127135050715879, 0.72602427797392, 0.841455427696928, 0.196573291206732, 0.825757328420877, 0.841111233457923, 0.922711052000523, 0.761531762313098, 0.825742032378912, 0.296645707683638, 0.620617144508287, 0.958264390705153, 0.307863801019266, 0.407151504652575, 0.282243936322629, 0.549482424044982, 0.713972105644643, 0.098097164882347, 0.537111163837835, 0.542020070366561, 0.176807268057019, 0.353880507638678, 0.489777490496635, 0.790795183507726, 0.0592662733979523, 0.459028488025069, 0.151738557033241, 0.0929551594890654, 0.0779134132899344, 0.307148020947352, 0.546055246377364, 0.837016093311831, 0.171284773619846, 0.587575771845877, 0.241601114859805, 0.31337141781114, 0.167342965025455, 0.0685987512115389, 0.669066821457818, 0.954979687929153, 0.332549912855029, 0.566485314397141, 0.263359896605834, 0.319177754689008, 0.339057768462226, 0.807553288294002, 0.520527699496597, 0.811745502986014, 0.561482779448852, 0.647233354859054, 0.34342527622357, 0.903135844040662, 0.854339107172564, 0.918922569602728, 0.0955074087250978, 0.387866671895608, 0.0487377096433192, 0.571297793881968, 0.656839879229665, 0.205045707756653, 0.0625816013198346, 0.78345129522495, 0.12332188943401, 0.0656368182972074, 0.809253866085783, 0.296993083320558, 0.642383235273883, 0.973497646395117, 0.174109410028905, 0.953753942158073, 0.259033127455041, 0.705334688303992, 0.0984951676800847, 0.691465334966779, 0.4406349780038, 0.599747243337333, 0.333237962564453, 0.141227112850174, 0.418277078075334, 0.156535034067929, 0.0130117551889271, 0.678242780733854, 0.364900819025934, 0.642083673039451, 0.308878709329292, 0.773856882937253, 0.344008211744949, 0.0298149816226214, 0.478354566730559, 0.173195263603702, 0.482871545478702, 0.667397376149893, 0.106602536514401, 0.13999799429439, 0.257256826385856, 0.352989153238013, 0.554053991334513, 0.890949654625729, 0.0977544717025012, 0.870101779000834, 0.00641228933818638, 0.0425303913652897, 0.80213169218041, 0.796984553337097, 0.838943901937455, 0.973829264519736, 0.579577002441511, 0.322493747342378, 0.177999842911959, 0.125597243895754, 0.658786106389016, 0.870640474371612, 0.640088250627741, 0.741533171618357, 0.66371796419844, 0.68093150947243, 0.100551261333749, 0.948710246942937, 0.385622742120177, 0.479253817116842, 0.115188874071464, 0.966730256332085, 0.280244067311287, 0.208170181140304, 0.440434583229944, 0.564603309379891, 0.0878979610279202, 0.183075766079128, 0.0284805046394467, 0.53812134033069, 0.344290125416592, 0.237841333029792, 0.944023978197947, 0.333733208710328, 0.307007843628526, 0.955937126185745, 0.251733649056405, 0.373507772572339, 0.155334897805005, 0.669178789481521, 0.615923454519361, 0.840029838727787, 0.99374789558351, 0.688335139770061, 0.671854331390932, 0.73317538597621, 0.284029302885756, 0.532713778549805, 0.538374504540116, 0.0989244049414992, 0.686416414333507, 0.293291210429743, 0.481543868780136, 0.516009592451155, 0.0879886180628091, 0.167972203111276, 0.329677867470309, 0.919222655240446, 0.188638334162533, 0.97517346823588, 0.0483143718447536, 0.82865811069496, 0.390194527572021, 0.456229547504336, 0.948849933687598, 0.643069158308208, 0.401304746046662, 0.878912170883268, 0.0822725496254861, 0.168719836976379, 0.381881823996082, 0.656038538552821, 0.775753572350368, 0.408817141084, 0.415541404858232, 0.318150263512507, 0.291011830093339, 0.182387807406485, 0.195520344655961, 0.00976702058687806, 0.497446309542283, 0.378525553504005, 0.336582147981972, 0.0113241064827889, 0.232298353919759, 0.289599416079, 0.648857319960371, 0.706553014693782, 0.328346054535359, 0.946232492569834, 0.971842027967796, 0.443784489063546, 0.980114311445504, 0.299686236307025, 0.126311125233769, 0.535374319180846, 0.715775988763198, 0.797569000627846, 0.683227876201272, 0.552273565437645, 0.226329773431644, 0.905094580492005, 0.816394352819771, 0.957277230219916, 0.893357448745519, 0.47008078196086, 0.354242675937712, 0.721320367185399, 0.0925244812387973, 0.738777130609378, 0.582541850628331, 0.484319257317111, 0.0803933788556606, 0.76493495493196, 0.0561430046800524, 0.0254009552299976, 0.593721231678501, 0.113005834165961, 0.543369693681598, 0.674849111121148, 0.769322903128341, 0.911817533662543, 0.991625806316733, 0.122617758344859, 0.488771989243105, 0.755433831829578, 0.769859830616042, 0.583585869520903, 0.607002418488264, 0.296526406193152, 0.351025643991306, 0.776102846488357, 0.239750403910875, 0.984474634984508, 0.750818685395643, 0.807890719501302, 0.0105472356081009, 0.431044669821858, 0.378955878317356, 0.711219529388472, 0.0251205959357321, 0.788759714225307, 0.512067185947672, 0.162757432553917, 0.620229818159714, 0.335608479101211, 0.483242811867967, 0.134625839302316, 0.569041446782649, 0.786967179039493, 0.973143812501803, 0.751344136428088, 0.304618034046143, 0.333127286750823, 0.184321743901819, 0.30129774985835, 0.47765814838931, 0.088481628568843, 0.338673058198765, 0.903101241914555, 0.929733412107453, 0.9421568994876, 0.898307794705033, 0.866225474514067, 0.0668448170181364, 0.0885649344418198, 0.640196219552308, 0.838498539058492, 0.519765058299527, 0.0402233400382102, 0.131482636556029, 0.465649048564956, 0.0249568945728242, 0.642110977321863, 0.987829617457464, 0.416865939740092, 0.950769182061777, 0.599921070272103, 0.691335887182504, 0.716418253723532, 0.349749396089464, 0.0604306855238974, 0.912198907695711, 0.122756147990003, 0.458381018135697, 0.308826263761148, 0.848479482345283, 0.587727086618543, 0.829343384830281, 0.0083794635720551, 0.883736587828025, 0.379561752546579, 0.869318788172677, 0.704269186593592, 0.116993663134053, 0.289279101183638, 0.957023909082636, 0.808971205027774, 0.128832846181467, 0.586152577772737, 0.643954464234412, 0.151724313152954, 0.158593793399632, 0.190588374156505, 0.951008136384189, 0.331976138986647, 0.330861279508099, 0.632964235730469, 0.780479865381494, 0.539334098109975, 0.339963287347928, 0.169603475136682, 0.661845166003332, 0.575142447138205, 0.774201167281717, 0.376458402955905, 0.427220516605303, 0.862489687046036, 0.823628956684843, 0.804404074791819, 0.465205915505067, 0.324450241634622, 0.630168482661247, 0.133318502223119, 0.91916462732479, 0.538245311239734, 0.502848829841241, 0.0272478316910565, 0.774237082106993, 0.437295780982822, 0.611940526636317, 0.719031858956441, 0.243023483548313, 0.507289197063074, 0.308967935852706, 0.645900930278003, 0.177874107612297, 0.556983257643878, 0.593390868976712, 0.368206968531013, 0.360891688615084, 0.836185275111347, 0.0361815229989588, 0.574729830957949, 0.246888515073806, 0.30530678993091, 0.508644068380818, 0.93962570396252, 0.150674900505692, 0.0120605167467147, 0.294344563502818, 0.373781717848033, 0.518686824478209, 0.447836559033021, 0.697079137898982, 0.665988822467625, 0.489891881588846, 0.667852543992922, 0.411332872696221, 0.526286926819012, 0.876767757115886, 0.350005661370233, 0.287977014901116, 0.123009173898026, 0.710778023349121, 0.401417459594086, 0.450388056226075, 0.507691232487559, 0.185676732333377, 0.96840141271241, 0.32310383906588, 0.276790167205036, 0.225533611373976, 0.513807603856549, 0.128874903079122, 0.0783426312264055, 0.133809349033982, 0.549861898878589, 0.418877564603463, 0.402486099628732, 0.449669334571809, 0.63839075807482, 0.653443341841921, 0.177323802839965, 0.153683312237263, 0.100016645854339, 0.203884218120947, 0.946000877069309, 0.326312511228025, 0.633370535913855, 0.938466720283031, 0.100801792461425, 0.422343526501209, 0.962828722316772, 0.917248809011653, 0.461132159922272, 0.883734453236684, 0.472966537578031, 0.707568701123819, 0.872352522565052, 0.580849524121732, 0.878127756994218, 0.791292174253613, 0.0183178004808724, 0.22283793007955, 0.618551770690829, 0.159674501279369, 0.615630849497393, 0.946554575813934, 0.243164661340415, 0.930276649305597, 0.857103840215132, 0.561269069788978, 0.499037636443973, 0.0670560311991721, 0.993084339657798, 0.208585757529363, 0.392882108921185, 0.0352825550362468, 0.999560774071142, 0.321793943177909, 0.56725717894733, 0.337137259775773, 0.348168013850227, 0.0510740370955318, 0.964910569600761, 0.335557013517246, 0.513903019716963, 0.0482541674282402, 0.432811581762508, 0.634976480156183, 0.998650589026511, 0.884697635890916, 0.606662029167637), actual = c(1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.64133731671609, 0.929182791616768, 0.184422378893942, 0.66538821789436, 0.748716862406582, 0.825391141697764, 0.167199081974104, 0.981051343958825, 0.58592470921576, 0.889017500681803, 0.888957901159301, 0.136870119255036, 0.186513961758465, 0.234839627053589, 0.597294124541804, 0.34362934762612, 0.594171432079747, 0.176719649927691, 0.427476950688288, 0.288009956944734, 0.722207258455455, 0.334331960650161, 0.131449408829212, 0.530833048280329, 0.363219454884529, 0.526054922724143, 0.641515915514901, 0.245207826606929, 0.876838876632974, 0.442939593223855, 0.971751526929438, 0.616711071692407, 0.831319689983502, 0.411642617546022, 0.909762225812301, 0.00656047649681568, 0.106058263219893, 0.794685752829537, 0.311981735285372, 0.494688550475985, 0.421502420445904, 0.890902895014733, 0.967490997863933, 0.0707353232428432, 0.211879851063713, 0.837434435961768, 0.031173383584246, 0.0684316789265722, 0.19618553458713, 0.356304636457935, 0.748182084178552, 0.942877408582717, 0.0929993602912873, 0.553516635904089, 0.0859778341837227, 0.833853557705879, 0.17463265103288, 0.512832646258175, 0.941470133606344, 0.311909441137686, 0.840318861883134, 0.986104722134769, 0.7210381582845, 0.701284213224426, 0.968395001254976, 0.908203331520781, 0.923691084841266, 0.00219037989154458, 0.508410212816671, 0.251707413000986, 0.215462261810899, 0.759844530373812, 0.463225632207468, 0.894798820838332, 0.998048005159944, 0.821868903702125, 0.531946763629094, 0.218322461703792, 0.669199240859598, 0.866880419431254, 0.236876304494217, 0.485493248095736, 0.700193987693638, 0.309505653800443, 0.392234032275155, 0.223531145369634, 0.84977374272421, 0.568446859717369, 0.555319901788607, 0.867085336241871, 0.805649433517829, 0.314167681382969, 0.675807813648134, 0.415243736933917, 0.126348194666207, 0.453412628732622, 0.243467407301068, 0.664878592593595, 0.229267292888835, 0.141965237213299, 0.702718075830489, 0.0565967501606792, 0.314295118208975, 0.517883985070512, 0.444517160765827, 0.582873763516545, 0.549801629735157, 0.448347724741325, 0.86050045164302, 0.600641296012327, 0.792088519781828, 0.310323834652081, 0.154032511403784, 0.344147445866838, 0.543774263234809, 0.402348997537047, 0.534282140200958, 0.362339067040011, 0.47798552759923, 0.310056109679863, 0.397443177644163, 0.369662613142282, 0.7225695787929, 0.880808954127133, 0.935649271821603, 0.575264266924933, 0.286711231106892, 0.39074489613995, 0.715386972529814, 0.380952537292615, 0.785104673588648, 0.95167372119613, 0.408785372273996, 0.188141906401142, 0.808260092046112, 0.775921571301296, 0.610170257044956, 0.325586228165776, 0.276797686470672, 0.631078049307689, 0.302513416158035, 0.248411137610674, 0.0173134848009795, 0.548755577765405, 0.0834747334010899, 0.151604024460539, 0.43758876575157, 0.00101934536360204, 0.0196100405883044, 0.0843688712920994, 0.871502287685871, 0.491110821021721, 0.970516594825312, 0.500348741188645, 0.854319023899734, 0.717068032128736, 0.45009095268324, 0.846333915367723, 0.417442713864148, 0.127765122335404, 0.0319229278247803, 0.197543918155134, 0.677908389130607, 0.653023679275066, 0.288085773587227, 0.475517863873392, 0.242300069192424, 0.642217597924173, 0.193065717816353, 0.908680200809613, 0.325155173195526, 0.035015138797462, 0.0279988781549037, 0.0812412307132035, 0.373707971768454, 0.271646012784913, 0.409118067473173, 0.874159740982577, 0.686954928794876, 0.299782793270424, 0.69746444793418, 0.728252345463261, 0.491024586372077, 0.47365049039945, 0.357867003185675, 0.205387151800096, 0.427112036151811, 0.23316948977299, 0.348620543489233, 0.343947341898456, 0.909217943437397, 0.277097741141915, 0.686524803750217, 0.0278383209370077, 0.255075143883005, 0.543970296392217, 0.972809240687639, 0.626782301114872, 0.171898954547942, 0.888139258138835, 0.948226978536695, 0.924503711750731, 0.434502922231331, 0.441502691013739, 0.874311984051019, 0.240245765075088, 0.578325627138838, 0.988205147907138, 0.989245726261288, 0.999002920696512, 0.518172356765717, 0.0466098268516362, 0.793591110035777, 0.235047538066283, 0.622358517721295, 0.0314197118859738, 0.113552285125479, 0.836103543173522, 0.901395086664706, 0.86937132361345, 0.417767728213221, 0.800289676524699, 0.503195420606062, 0.963500304380432, 0.745156874181703, 0.0625688964501023, 0.0335946383420378, 0.0294598361942917, 0.373973027803004, 0.519792023580521, 0.788974299328402, 0.213230153545737, 0.201237073168159, 0.23761444282718, 0.833548647118732, 0.493093661265448, 0.133432015776634, 0.360866947332397, 0.973124312702566, 0.920930664986372, 0.18508653412573, 0.712172738742083, 0.958070103777573, 0.812852239934728, 0.65801750542596, 0.319624966243282, 0.0577583843842149, 0.363145073410124, 0.811578302178532, 0.558661744464189, 0.931085132993758, 0.861573351314291, 0.29334549466148, 0.198724525049329, 0.452562993625179, 0.0556172789074481, 0.763155844761059, 0.87777573405765, 0.464507268043235, 0.791393546853215, 0.0826098721008748, 0.63885788875632, 0.235181517666206, 0.138458604458719, 0.19050857773982, 0.0561847104690969, 0.645878561306745, 0.36103348294273, 0.702861990779638, 0.394613746087998, 0.429452911484987, 0.240640240255743, 0.0610163076780736, 0.973487825598568, 0.0541986266616732, 0.943669314263389, 0.244718827540055, 0.506786319892853, 0.993430491536856, 0.27187465201132, 0.615106358192861, 0.682750017382205, 0.822815658058971, 0.9664979246445, 0.7531360052526, 0.711877934169024, 0.936272396007553, 0.901188924908638, 0.795553828822449, 0.350689262151718, 0.207743825390935, 0.246028307126835, 0.10741148586385, 0.591386356856674, 0.895028916420415, 0.621452478924766, 0.924249751493335, 0.499455593526363, 0.0880907285027206, 0.408270090818405, 0.00585115049034357, 0.993026670534164, 0.800935119623318, 0.395456483354792, 0.77342640212737, 0.0180015137884766, 0.488942480413243, 0.756290618795902, 0.596031713066623, 0.0704560277517885, 0.9428092809394, 0.9116585040465, 0.747441828018054, 0.887933046324179, 0.971407774137333, 0.827012083493173, 0.634321766439825, 0.848974504740909, 0.163196542067453, 0.583751146448776, 0.220831249607727, 0.250151274027303, 0.0332394195720553, 0.770590655272827, 0.414062391733751, 0.0342664900235832, 0.867739935638383, 0.855531178181991, 0.731469280086458, 0.0530336829833686, 0.183032173430547, 0.22164523624815, 0.393863240722567, 0.782943835249171, 0.93113008281216, 0.0687044190708548, 0.535146965179592, 0.789090044097975, 0.2800368135795, 0.460800196044147, 0.52870203461498, 0.514386928640306, 0.0559230879880488, 0.102681192569435, 0.67371557909064, 0.577585091348737, 0.899331468390301, 0.871389521285892, 0.796672465279698, 0.306394402403384, 0.357579755131155, 0.78226914559491, 0.829249443951994, 0.872613694751635, 0.256439733318985, 0.053222878370434, 0.8229953518603, 0.0508415547665209, 0.800988449715078, 0.658850803272799, 0.483561467379332, 0.88215879490599, 0.186427862383425, 0.468891273951158, 0.907150686252862, 0.109814532799646, 0.355138101847842, 0.992612664122134, 0.135979897808284, 0.0829708059318364, 0.830541084753349, 0.303159796632826, 0.633254293818027, 0.640720122726634, 0.801624804968014, 0.44799781544134, 0.541919811163098, 0.961906400509179, 0.134060650365427, 0.793847872875631, 0.483033762080595, 0.791320753516629, 0.659092140384018, 0.176775031257421, 0.839050516486168, 0.148357508238405, 0.908596832770854, 0.499093969352543, 0.912672798382118, 0.386151939164847, 0.696624527219683, 0.821893846383318, 0.82873102882877, 0.126217376207933, 0.367793556069955, 0.193841965869069, 0.718233559979126, 0.941820348845795, 0.980987544404343, 0.821445055305958, 0.231604773318395, 0.323017939459532, 0.264648076146841, 0.346379245398566, 0.213191435206681, 0.00149380858056247, 0.239353721961379, 0.47474595811218, 0.149696222972125, 0.655747265322134, 0.665622952161357, 0.134613192407414, 0.475118623580784, 0.838868906255811, 0.109681688481942, 0.329188638133928, 0.641641630092636, 0.296516997972503, 0.19178027077578, 0.626497281715274, 0.748320130398497, 0.823116946732625, 0.835525620961562, 0.938902511261404, 0.191079350188375, 0.548214202513918, 0.173361405031756, 0.601355471415445, 0.437960421433672, 0.848596593132243, 0.821918980916962, 0.327416566899046, 0.439322800608352, 0.0107888600323349, 0.388003202388063, 0.204287027707323, 0.0878553092479706, 0.448151115095243, 0.263108829967678, 0.195187645498663, 0.833327211905271, 0.379851332865655, 0.971631634281948, 0.729458196088672, 0.332101901061833, 0.712186715798452, 0.724544922821224, 0.491846980759874, 0.140778281260282, 0.42055531963706, 0.248767551966012, 0.23760493658483, 0.106623757630587, 0.0961456887889653, 0.576374265830964, 0.957963897148147, 0.84900734317489, 0.344289910281077, 0.867749485652894, 0.850809583906084, 0.333482839632779, 0.887015039566904, 0.0999821012374014, 0.0672601223923266, 0.318809646647424, 0.480259814532474, 0.154785158112645, 0.6678265130613, 0.35648802947253, 0.840358428424224, 0.957356518367305, 0.445950065273792, 0.24136974522844, 0.412743177264929, 0.814724993426353, 0.895316534908488, 0.750459891976789, 0.759650141233578, 0.409005867084488, 0.0747935478575528, 0.226535219233483, 0.835848512593657, 0.789265815168619, 0.113752698758617, 0.902205536141992, 0.883535866625607, 0.471280438592657, 0.24369430472143, 0.375551536213607, 0.772123072762042, 0.266897590365261, 0.197845326736569, 0.738883554935455, 0.0703341895714402, 0.866743069374934, 0.923335809260607, 0.344603540608659, 0.479330590227619, 0.614121312974021, 0.293896046234295, 0.00490327645093203, 0.454961819108576, 0.750631225295365, 0.580332365818322, 0.199099406134337, 0.693458684021607, 0.848335473099723, 0.173562918324023, 0.548492085188627, 0.64617598708719, 0.605995082994923, 0.773167388979346, 0.817340809851885, 0.248959051910788, 0.351852473570034, 0.539847800740972, 0.453785160789266, 0.484220891259611, 0.262044747360051, 0.475075478199869, 0.851306622382253, 0.27174395066686, 0.174358672928065, 0.507842851104215, 0.847417487064376, 0.887149591464549, 0.999093937687576, 0.269528464879841, 0.919251560000703, 0.781411670614034, 0.901284847641364, 0.678510818397626, 0.907457365421578, 0.985453309258446, 0.18252802407369, 0.235140566248447, 0.64600958651863, 0.837801919085905, 0.0997383864596486, 0.598388922633603, 0.683102080831304, 0.651584244798869, 0.964297157712281, 0.295579680940136, 0.762369652744383, 0.65328511968255, 0.632235612021759, 0.914794174022973, 0.983894457109272, 0.376180981984362, 0.490259703015909, 0.431214719079435, 0.485516316723078, 0.911671499721706, 0.64360416540876, 0.5251588774845, 0.240425835596398, 0.606991658220068, 0.565877877408639, 0.621595815289766, 0.483381147962064, 0.473929313709959, 0.530055164825171, 0.871047255583107, 0.854723969474435, 0.975535712204874, 0.667569032171741, 0.816408487968147, 0.711832383647561, 0.581615063361824, 0.433226689696312, 0.157969768857583, 0.46238873526454, 0.527122375788167, 0.744400138035417, 0.873182361479849, 0.243384390836582, 0.476732673123479, 0.413037993246689, 0.48098835721612, 0.431321948068216, 0.0860029025934637, 0.78413164592348, 0.198816117364913, 0.607920970767736, 0.139274925459176, 0.962623762432486, 0.918159916298464, 0.00457954430021346, 0.054376277141273, 0.881259705405682, 0.102634051349014, 0.893787485314533, 0.866372557124123, 0.87030946998857, 0.97310662525706, 0.19691520021297, 0.130589082138613, 0.47876140777953, 0.952091926708817, 0.84165309346281, 0.518611440435052, 0.859307456528768, 0.868573665386066, 0.605168397538364, 0.889400842599571, 0.75638190144673, 0.692510223016143, 0.370729196583852, 0.254920857492834, 0.110029165167361, 0.369631048757583, 0.326198065886274, 0.46897700545378, 0.816170149715617, 0.402325537987053, 0.460513406200334, 0.660212498623878, 0.830694837030023, 0.975083795841783, 0.149838278768584, 0.842733095865697, 0.995229320600629, 0.412153843790293, 0.626346168341115, 0.274637616006657, 0.431917321169749, 0.348370634019375, 0.225963247939944, 0.697394306305796, 0.96433147508651, 0.0864987627137452, 0.522923437412828, 0.725308121414855, 0.372029928257689, 0.541506685083732, 0.637971576303244, 0.0105213185306638, 0.894789342069998, 0.938760594232008, 0.748738601105288, 0.996789966477081, 0.574482418596745, 0.256294415332377, 0.321106670191512, 0.249579672934487, 0.642406646860763, 0.908306554891169, 0.194274893030524, 0.0907926955260336, 0.720729324268177, 0.00577536015771329, 0.176426958292723, 0.0728080119006336, 0.410328585421667, 0.0542944953776896, 0.941461068345234, 0.64570349175483, 0.0638334245886654, 0.138385434867814, 0.445548882475123, 0.557474998757243, 0.694911836646497, 0.96409514779225, 0.839456764049828, 0.0243889056146145, 0.113682885654271, 0.676886998349801, 0.921547312289476, 0.388613671064377, 0.677903177915141, 0.283716855570674, 0.181516043376178, 0.795637286966667, 0.938382038846612, 0.87867558025755, 0.153908247128129, 0.110506896860898, 0.9119989995379, 0.670321814948693, 0.122201974270865, 0.614853453123942, 0.412762533640489, 0.18162807635963, 0.96191071299836, 0.552308870945126, 0.830460031284019, 0.277131809387356, 0.195376859046519, 0.216909252805635, 0.112518580630422, 0.279640807537362, 0.383817573310807, 0.110779100563377, 0.147010774351656, 0.761816639686003, 0.377141600940377, 0.60555725521408, 0.0659778050612658, 0.00999182043597102, 0.110834267921746, 0.191227172268555, 0.903852554270998, 0.979958394309506, 0.720061914995313, 0.222974308300763, 0.489811222767457, 0.496435551438481, 0.861032794928178, 0.722039405023679, 0.122218598844483, 0.774318318581209, 0.669665533816442, 0.0582697619684041, 0.437672044150531, 0.0680408368352801, 0.580292179482058, 0.0395415360108018, 0.709950390039012, 0.289525265106931, 0.473671440733597, 0.846561939222738, 0.447378893615678, 0.203500899951905, 0.377166534541175, 0.560389592777938, 0.455781903350726, 0.716117940610275, 0.796465067658573, 0.171660783700645, 0.1234920269344, 0.542308853240684, 0.278147063450888, 0.356937863631174, 0.767526414478198, 0.0854493782389909, 0.787933225743473, 0.327058316441253, 0.646756259957328, 0.0998457993846387, 0.580830543302, 0.280414893990383, 0.867643468081951, 0.977897570468485, 0.357131469761953, 0.591216497588903, 0.92488181963563, 0.317757305921987, 0.916224329965189, 0.00461145886220038, 0.710373893380165, 0.961158635560423, 0.470081519335508, 0.415986607782543, 0.761854307027534, 0.853547194041312, 0.228589249541983, 0.400189842330292, 0.372079506749287, 0.717005945043638, 0.435283113736659, 0.305828090989962, 0.684967258479446, 0.622127415146679, 0.289748060517013, 0.641072248341516, 0.282838540850207, 0.410028791287914, 0.256548072444275, 0.998804371803999, 0.938919183099642, 0.624298317590728, 0.343038453487679, 0.570341684157029, 0.667911149095744, 0.617284843232483, 0.606611857423559, 0.154115403769538, 0.501587220700458, 0.282973747467622, 0.241213629953563, 0.239260428119451, 0.244019910460338, 0.0298611705657095, 0.764264357276261, 0.57359106419608, 0.565226773265749, 0.884924289770424, 0.0450447374023497, 0.866142375627533, 0.445406384766102, 0.865370183484629, 0.678707332815975, 0.378717078594491, 0.744026542408392, 0.211575438268483, 0.169593388447538, 0.264882389688864, 0.611096826381981, 0.376498757163063, 0.717478049686179, 0.752133175730705, 0.939821174833924, 0.793437772197649, 0.26975493086502, 0.630782107124105, 0.023372219176963, 0.0844521489925683, 0.343504363903776, 0.923828057246283, 0.306794100441039, 0.214814518578351, 0.0505336963105947, 0.0977063574828207, 0.639023382682353, 0.66716488962993, 0.643543256213889, 0.246059369994327, 0.0331643142271787, 0.00757245626300573, 0.985645152628422, 0.0696925080846995, 0.688611010555178, 0.557258938672021, 0.620008002733812, 0.887956985039636, 0.427651327336207, 0.350818135542795, 0.864012386417016, 0.688962941290811, 0.491766010178253, 0.91755608189851, 0.640690281055868, 0.758428572444245, 0.35622368240729, 0.821199991274625, 0.431850097374991, 0.000101050129160285, 0.196163546992466, 0.499919401248917, 0.260597777785733, 0.14950468018651, 0.234607228077948, 0.2569273582194, 0.483188280602917, 0.902590508107096, 0.292561436304823, 0.412288467632607, 0.877903721295297, 0.864736107178032, 0.37539971107617, 0.952365280361846, 0.763764160685241, 0.184247193625197, 0.567463728832081, 0.56672435044311, 0.827181773260236, 0.222659200197086, 0.855334352701902, 0.240882474463433, 0.0979399054776877, 0.199997511226684, 0.58335870760493, 0.180105665232986, 0.343321598600596, 0.812391521874815, 0.360162160592154, 0.639648975804448, 0.189172310056165, 0.260609785793349, 0.457995718577877, 0.904545221943408, 0.709744008257985, 0.693564121611416, 0.586143249180168, 0.324625689070672, 0.484423313988373, 0.344033768167719, 0.191113729262725, 0.70558954984881, 0.064532230142504, 0.306871270760894, 0.0827734116464853, 0.49163028318435, 0.535507723456249, 0.621966270031407, 0.0121379687916487, 0.27369013777934, 0.202873209491372, 0.499194414354861, 0.794543339172378, 0.190485833212733, 0.235996919451281, 0.834189440589398, 0.438171516638249, 0.0276203432586044, 0.949699147837237, 0.960974727524444, 0.142902194289491, 0.213023374788463, 0.562213031109422, 0.765785162802786, 0.0868696942925453, 0.156142679508775, 0.99330261698924, 0.405733541119844, 0.345882354304194, 0.806003499077633, 0.805907224072143, 0.340800980571657, 0.626983089372516, 0.422113033011556, 0.892775040352717, 0.558418164262548, 0.438621613429859, 0.670733829727396, 0.445151971885934, 0.851510991109535, 0.539426076225936, 0.276782973669469, 0.842976870248094, 0.505127845099196, 0.148100984049961, 0.136715258238837, 0.884339969838038, 0.377059242455289, 0.0431962762959301, 0.386904476210475, 0.954140918562189, 0.932960958452895, 0.568512365454808, 0.887756495503709, 0.740278284996748, 0.682074235985056, 0.803219225024804, 0.909194696694613, 0.116553807398304, 0.128291703993455, 0.462006310001016, 0.502352586714551, 0.74179064319469, 0.00434423983097076, 0.768223898019642, 0.364923094632104, 0.460373842855915, 0.292124150553718, 0.504886370152235, 0.963357479544356, 0.291618882445619, 0.8485092590563, 0.614992667455226, 0.0221232802141458, 0.833714503096417, 0.029018595116213, 0.396063105436042, 0.66207472095266, 0.185533510753885, 0.339935453375801, 0.580149297369644, 0.753082685172558, 0.0347974910400808, 0.52543301647529, 0.644806925207376, 0.935075650922954, 0.00992296310141683, 0.68265767255798, 0.177261856850237, 0.227050703950226, 0.0568041182123125, 0.641924685332924, 0.0192301191855222, 0.287557557225227, 0.406069624470547, 0.918059691088274, 0.801776514388621, 0.668253801530227, 0.266078690299764, 0.247022785712034, 0.329066702630371, 0.753670135512948, 0.168501661159098, 0.474315161351115, 0.613525959430262, 0.389433688251302, 0.896053211996332, 0.437318166950718, 0.599643664434552, 0.302485740976408, 0.0582327251322567, 0.265292898984626, 0.837229564785957, 0.362429582281038, 0.352254339959472, 0.988144392380491, 0.823401605477557, 0.172417122172192, 0.0569532965309918, 0.708707592217252, 0.426036958349869, 0.914329871768132), actual = c(0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"