Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4857125        0.4866385 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4866385        0.4866385        0.5333747 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.521668101893738, 0.586188117740676, 0.184896775521338, 0.808074223110452, 0.800097777042538, 0.855760673061013, 0.570392798865214, 0.971732439706102, 0.0706696636043489, 0.786423495272174, 0.3001358427573, 0.533357128500938, 0.946611354127526, 0.120329380268231, 0.622138257371262, 0.0426072487607598, 0.293858769815415, 0.680015957448632, 0.817183873383328, 0.263432984706014, 0.780018492368981, 0.0350915351882577, 0.998214933555573, 0.494445935357362, 0.238673309562728, 0.802988843992352, 0.876981874229386, 0.0490406909957528, 0.774259004741907, 0.388011164730415, 0.781612888909876, 0.197282850276679, 0.0376884352881461, 0.845910650445148, 0.957796770613641, 0.253812793875113, 0.846016869647428, 0.64982649916783, 0.815661327214912, 0.27981984638609, 0.166275684721768, 0.00102341920137405, 0.591210601385683, 0.220376384444535, 0.573375195031986, 0.527395784622058, 0.193897993536666, 0.50690735061653, 0.252399751450866, 0.316381342476234, 0.0182651639916003, 0.37745888507925, 0.471774955978617, 0.876452340744436, 0.710982297779992, 0.734375053551048, 0.807052862597629, 0.912313436158001, 0.973058146191761, 0.258620823500678, 0.382574816001579, 0.797495493898168, 0.00250625074841082, 0.665302446344867, 0.706924109719694, 0.0982822941150516, 0.808929758379236, 0.581852553412318, 0.190449095796794, 0.68558823922649, 0.231780167901888, 0.707391645759344, 0.954442107817158, 0.969504144508392, 0.522064396180212, 0.47075094236061, 0.393184005981311, 0.628580007469282, 0.152874036226422, 0.28690008725971, 0.766626553377137, 0.621147928992286, 0.843017010716721, 0.632105756318197, 0.392146085621789, 0.848858341341838, 0.100632786750793, 0.312715268693864, 0.919680064776912, 0.519629688933492, 0.481505728326738, 0.660203871782869, 0.806997780222446, 0.918373117456213, 0.248746653087437, 0.768027376383543, 0.785465480526909, 0.609484708867967, 0.494547917274758, 0.670015714131296, 0.658885781187564, 0.140582570806146, 0.507342533674091, 0.34168934286572, 0.16815907577984, 0.316128827864304, 0.122241341276094, 0.967126893112436, 0.0347759642172605, 0.890797830419615, 0.295925484970212, 0.465422368841246, 0.167806573444977, 0.0290313975419849, 0.882252075942233, 0.448155308840796, 0.492365762824193, 0.181263587204739, 0.937321363715455, 0.86482820357196, 0.0213831015862525, 0.905285621294752, 0.617400763556361, 0.836652752244845, 0.401517143240198, 0.651235425379127, 0.339450403349474, 0.0713948144111782, 0.41944853332825, 0.705319737782702, 0.788979325909168, 0.309145516715944, 0.802760691614822, 0.269528297474608, 0.530737278051674, 0.120351848425344, 0.459487709216774, 0.226190830115229, 0.460463814204559, 0.0502875843085349, 0.119521803921089, 0.0183308213017881, 0.202113633975387, 0.92841005185619, 0.10216787038371, 0.0104888104833663, 0.668686978518963, 0.722635387908667, 0.723290100926533, 0.244529296644032, 0.707319928333163, 0.630259053083137, 0.252247470896691, 0.17479033395648, 0.139003991847858, 0.0508046252653003, 0.510601581772789, 0.247919146902859, 0.166300992714241, 0.604874158510938, 0.895399758592248, 0.492058385862038, 0.779963531065732, 0.477692471584305, 0.514395589008927, 0.138918654993176, 0.752067395020276, 0.385323236696422, 0.395111574325711, 0.405821990454569, 0.0268871234729886, 0.480936346109957, 0.980410360498354, 0.296146224718541, 0.755915583111346, 0.244891129201278, 0.16965302452445, 0.816468343837187, 0.839948098408058, 0.0593196076806635, 0.402365588117391, 0.775645165937021, 0.380758911604062, 0.839723902288824, 0.171099173370749, 0.776661061448976, 0.60951844137162, 0.617488381685689, 0.939808502327651, 0.339670786634088, 0.691000915830955, 0.849853650899604, 0.139238625066355, 0.692362299887463, 0.385771481785923, 0.0053265206515789, 0.38884553546086, 0.316429275088012, 0.729363515973091, 0.782483850140125, 0.616035524988547, 0.68318206188269, 0.71600799378939, 0.780349691165611, 0.81431769579649, 0.842919533839449, 0.391812486574054, 0.523091674316674, 0.541183524765074, 0.449337405851111, 0.660143584711477, 0.928276462014765, 0.339866950409487, 0.833456203574315, 0.907621774356812, 0.369515953352675, 0.364141466794536, 0.741965398658067, 0.583106977632269, 0.546783111058176, 0.966943155741319, 0.865474412217736, 0.0294956117868423, 0.998004689346999, 0.556461537489668, 0.230250268010423, 0.37093970971182, 0.943433364154771, 0.617351674241945, 0.154101117281243, 0.471342990640551, 0.544778677867725, 0.857524626422673, 0.632560220081359, 0.332096859114245, 0.654590266756713, 0.226315124891698, 0.626823029480875, 0.33730573952198, 0.34437999734655, 0.47403559461236, 0.48272121977061, 0.18987624719739, 0.486651820596308, 0.675288100028411, 0.535550453700125, 0.0260770567692816, 0.91274808626622, 0.166900482261553, 0.551700106589124, 0.881599543383345, 0.387785444967449, 0.501043781172484, 0.0312825844157487, 0.747087390394881, 0.741740266326815, 0.375052916351706, 0.0815159289631993, 0.775306851835921, 0.425803592894226, 0.457861099625006, 0.0244556420948356, 0.348432909231633, 0.157499358989298, 0.292110730661079, 0.103531800210476, 0.925470757996663, 0.567787470761687, 0.248340313555673, 0.13673763605766, 0.0622335388325155, 0.664464677684009, 0.915155255468562, 0.369185276096687, 0.700724843190983, 0.753031187923625, 0.417740339878947, 0.619431406026706, 0.80583345820196, 0.503596103051677, 0.645282261772081, 0.571555571863428, 0.483447453705594, 0.0794923480134457, 0.354610783746466, 0.0897328502032906, 0.0679609028156847, 0.213700348278508, 0.396978048142046, 0.876724816858768, 0.474765731254593, 0.140200493624434, 0.649096039123833, 0.640608002431691, 0.286055653123185, 0.4593588297721, 0.993987077148631, 0.277322078822181, 0.233761816285551, 0.669803082942963, 0.156899021938443, 0.427648024400696, 0.138231207849458, 0.710324501153082, 0.741370209492743, 0.84866359597072, 0.886612168746069, 0.472791648469865, 0.846576730487868, 0.174755905522034, 0.490359009243548, 0.326823529787362, 0.265484691830352, 0.496272013522685, 0.81012811511755, 0.38142255670391, 0.82769613317214, 0.517508235061541, 0.609696075087413, 0.0912725939415395, 0.970606726594269, 0.143748302012682, 0.817643525078893, 0.82778188213706, 0.89536071382463, 0.0961915319785476, 0.909215863794088, 0.235226889373735, 0.70786047168076, 0.142268483759835, 0.653252977179363, 0.875691080000252, 0.778752409154549, 0.530473577324301, 0.637561819050461, 0.19612320093438, 0.748311080271378, 0.806267934851348, 0.942445184802637, 0.272451514843851, 0.816568560199812, 0.857205965090543, 0.8635658249259, 0.959355320548639, 0.0671104788780212, 0.62559466669336, 0.669399607228115, 0.994674190180376, 0.837326966458932, 0.548913951264694, 0.501774478703737, 0.563758620526642, 0.1722557307221, 0.734295496717095, 0.311470285989344, 0.269886547233909, 0.00191693007946014, 0.0090363291092217, 0.709309858968481, 0.0901706691365689, 0.733057566685602, 0.429678832180798, 0.0764576834626496, 0.688849610742182, 0.322722735116258, 0.903203295310959, 0.855429176706821, 0.246316649019718, 0.0942017922643572, 0.772653905441985, 0.0471563653554767, 0.17504771053791, 0.8929126446601, 0.397316023008898, 0.468807575060055, 0.783960079774261, 0.424789931857958, 0.643839593976736, 0.341826973250136, 0.0575522016733885, 0.405346036422998, 0.959771348861977, 0.969358504749835, 0.2740693311207, 0.191618410870433, 0.597900675144047, 0.787946712691337, 0.51786522148177, 0.261832707794383, 0.266758412355557, 0.442691223463044, 0.0191554564516991, 0.946019098628312, 0.47216834127903, 0.559805304510519, 0.397091360064223, 0.24601872311905, 0.472948591690511, 0.753149731317535, 0.266175364609808, 0.436225822428241, 0.0808364595286548, 0.166322463657707, 0.332032290752977, 0.793801059713587, 0.193141063442454, 0.811101795407012, 0.813649588497356, 0.48272585333325, 0.574150255648419, 0.860898353392258, 0.0047684230376035, 0.718901270069182, 0.859586366685107, 0.309895575977862, 0.372056638589129, 0.797317758435383, 0.608052639756352, 0.0659245611168444, 0.902809735620394, 0.927355372812599, 0.952097089262679, 0.320617589866742, 0.29867099924013, 0.255040615797043, 0.503167167073116, 0.780095956753939, 0.174711317522451, 0.896608493058011, 0.358452790183946, 0.117718653520569, 0.657770663034171, 0.134818830294535, 0.459648911142722, 0.274914090754464, 0.156564726261422, 0.0647974398452789, 0.415250678081065, 0.0556309109088033, 0.48084725276567, 0.465700428234413, 0.0834645226132125, 0.491723840823397, 0.878227649489418, 0.339510583085939, 0.572047467110679, 0.0411553697194904, 0.00907012028619647, 0.449861196801066, 0.502155853435397, 0.295382371172309, 0.0273759893607348, 0.790054792538285, 0.123682825593278, 0.569896965753287, 0.817992600612342, 0.866305411793292, 0.634378747781739, 0.0152798390481621, 0.699740116950125, 0.0338908182457089, 0.616500741336495, 0.0957355140708387, 0.353314421139657, 0.835254412377253, 0.58144972124137, 0.736216250574216, 0.00401954515837133, 0.162939368980005, 0.370840030256659, 0.668908735038713, 0.191435013897717, 0.108152215136215, 0.1350018363446, 0.545979339396581, 0.200076330685988, 0.128262059064582, 0.829185022041202, 0.204525014385581, 0.757039312738925, 0.354222129564732, 0.0793133568949997, 0.779245225246996, 0.205283984541893, 0.0177902695722878, 0.347878421656787, 0.567344058770686, 0.603031100472435, 0.898740599164739, 0.508727354928851, 0.794400486396626, 0.958339361706749, 0.593409350141883, 0.778143736533821, 0.713061821646988, 0.19855335704051, 0.831255741184577, 0.100193672114983, 0.944663865026087, 0.772551057627425, 0.135665303794667, 0.392504525836557, 0.46646103891544, 0.883100794162601, 0.181255203206092, 0.0572721341159195, 0.686048814328387, 0.632794834906235, 0.0292080254293978, 0.23993838718161, 0.490006279665977, 0.0512897931039333, 0.221285748062655, 0.928649431327358, 0.0459307539276779, 0.533396699465811, 0.756539820460603, 0.687807450536638, 0.380988175747916, 0.910468152491376, 0.66454484150745, 0.573889972409233, 0.171453949064016, 0.717918237205595, 0.0707046578172594, 0.950697023887187, 0.864912427496165, 0.424969686893746, 0.645960579393432, 0.0520551814697683, 0.862309720134363, 0.469479144318029, 0.931003046222031, 0.272208099020645, 0.103737177560106, 0.444366306765005, 0.0213966439478099, 0.356430997373536, 0.392916251439601, 0.876319922972471, 0.502755115507171, 0.901128493016586, 0.700054297223687, 0.134052168345079, 0.0267206830903888, 0.103148452239111, 0.655484338290989, 0.249406282324344, 0.964157851412892, 0.26805980107747, 0.356402941280976, 0.828580789500847, 0.303081328514963, 0.865333549678326, 0.0474147114437073, 0.976192273199558, 0.45141816791147, 0.953242649324238, 0.836657354841009, 0.509571117116138, 0.439022459788248, 0.202004982624203, 0.696597124449909, 0.484235873678699, 0.0845352031756192, 0.808128104778007, 0.912659630645066, 0.971746014431119, 0.490162153262645, 0.543641119496897, 0.884019902674481, 0.146424760343507, 0.825257412157953, 0.708591599948704, 0.235431681387126, 0.0179692567326128, 0.535784793784842, 0.324200294679031, 0.761727457633242, 0.407567899441347, 0.581730827689171, 0.39205749402754, 0.514162496430799, 0.508100485894829, 0.154027670156211, 0.0493344308342785, 0.456857978366315, 0.734053173335269, 0.213796142721549, 0.395402383059263, 0.798768805107102, 0.803788038901985, 0.830985729116946, 0.549361970974132, 0.701480712508783, 0.491020564222708, 0.268819219199941, 0.590457648504525, 0.927913883468136, 0.912619315320626, 0.22237710095942, 0.62349580321461, 0.513087626313791, 0.751560916891322, 0.0411332987714559, 0.0811365849804133, 0.602719067130238, 0.137263006530702, 0.398884621215984, 0.568386456463486, 0.087366605643183, 0.059399195946753, 0.569363631540909, 0.94829523214139, 0.839402405545115, 0.863282394595444, 0.176146979210898, 0.213054017629474, 0.16003734478727, 0.63792331609875, 0.04623226611875, 0.723322216188535, 0.220296940300614, 0.179345177486539, 0.266974131343886, 0.528607451822609, 0.576782972319052, 0.150684079155326, 0.0652065891772509, 0.703692844836041, 0.216676085488871, 0.938795737689361, 0.872849751496688, 0.671366657828912, 0.345363845350221, 0.812557748053223, 0.114333101548254, 0.684741689823568, 0.0497454204596579, 0.66264603449963, 0.376835254719481, 0.79588137473911, 0.345502549782395, 0.139694412937388, 0.916982876835391, 0.574968103552237, 0.854004197753966, 0.123032010858878, 0.186523840995505, 0.740052531473339, 0.243532189633697, 0.402009840821847, 0.932258340064436, 0.1917370907031, 0.0356711156200618, 0.0680503149051219, 0.810960366623476, 0.384580363053828, 0.27914824616164, 0.728797902353108, 0.226578531088307, 0.448247463908046, 0.900236135115847, 0.310512471478432, 0.00975919119082391, 0.879500288516283, 0.444557732669637, 0.678754805587232, 0.633225134806708, 0.431779919890687, 0.99638171470724, 0.489185280632228, 0.478830621112138, 0.999936715466902, 0.356895824428648, 0.6488029838074, 0.900191504973918, 0.237024372210726, 0.165794002823532, 0.842313653323799, 0.0316510004922748, 0.32911738101393, 0.82127351523377, 0.642566486494616, 0.652473467634991, 0.943377633346245, 0.346042742021382, 0.000961175188422203, 0.534314145101234, 0.606076270341873, 0.0165386749431491, 0.770616980735213, 0.989889575401321, 0.30433857277967, 0.54159830417484, 0.0908700211439282, 0.740765759255737, 0.327763635432348, 0.480846335180104, 0.606008431641385, 0.0191793558187783, 0.356286288006231, 0.44038218469359, 0.619264317909256, 0.700971136102453, 0.176701055141166, 0.683568301144987, 0.365253666881472, 0.0562118401285261, 0.435652984539047, 0.0780469987075776, 0.851906729396433, 0.582847909536213, 0.281980942934752, 0.931812327122316, 0.598525505745783, 0.624670159537345, 0.956397976726294, 0.987333811353892, 0.375911387847736, 0.474892417667434, 0.18699032953009, 0.53197600832209, 0.60242916434072, 0.694003582000732, 0.722526470199227, 0.125312825664878, 0.166189234238118, 0.719751146854833, 0.558586340164766, 0.759471775498241, 0.264348458964378, 0.792056451085955, 0.327692560851574, 0.0599528630264103, 0.883390174014494, 0.421316276304424, 0.243697635596618, 0.721760193584487, 0.866278334287927, 0.513308751396835, 0.9296653221827, 0.220063714077696, 0.29944814206101, 0.00182926445268095, 0.218272599391639, 0.419600289314985, 0.417230924125761, 0.0128475897945464, 0.484682414215058, 0.37487502489239, 0.134841086342931, 0.647096975240856, 0.52177832252346, 0.897952024359256, 0.882960588904098, 0.025859851622954, 0.113027968676761, 0.679913275176659, 0.547255701385438, 0.489094061078504, 0.621667470782995, 0.256667306879535, 0.971677316818386, 0.680661711841822, 0.906338471220806, 0.898947625886649, 0.64837071346119, 0.04052132088691, 0.80253844987601, 0.330752622336149, 0.436219661962241, 0.208969302941114, 0.144432822009549, 0.480611976003274, 0.940787193598226, 0.173873524647206, 0.0276848336216062, 0.523948882007971, 0.386835760204121, 0.0145196486264467, 0.218146138126031, 0.441102761076763, 0.873008702881634, 0.0396333064418286, 0.852825079578906, 0.007005495717749, 0.779269048245624, 0.26489896979183, 0.901823226129636, 0.240935850189999, 0.462717091897503, 0.677894963650033, 0.419617602135986, 0.538897863589227, 0.0398009973578155, 0.211510247318074, 0.819251052336767, 0.915049114031717, 0.730613163206726, 0.617247552843764, 0.646351744886488, 0.502613968215883, 0.407563176006079, 0.94993858365342, 0.868513860739768, 0.30907261534594, 0.00200303224846721, 0.302106045885012, 0.813753586495295, 0.489147268468514, 0.990084140561521, 0.561829816550016, 0.608998740790412, 0.897172461496666, 0.974103488493711, 0.051531670615077, 0.290477047208697, 0.586710483301431, 0.177542961435392, 0.636890269583091, 0.132252336246893, 0.99327310314402, 0.0397650958038867, 0.702916958136484, 0.148182987235487, 0.385113907046616, 0.208885160041973, 0.808218054706231, 0.500566014321521, 0.848045318620279, 0.785202413564548, 0.301367612788454, 0.673968051560223, 0.879473742563277, 0.322640742175281, 0.532380007905886, 0.292612091172487, 0.0518642850220203, 0.194387855473906, 0.0187237199861556, 0.492995796492323, 0.692033389117569, 0.852777231950313, 0.0926406336948276, 0.352888263761997, 0.794188007013872, 0.100630359724164, 0.103723111562431, 0.372843105345964, 0.172343205427751, 0.671104192268103, 0.274326909100637, 0.59284526668489, 0.84201125591062, 0.599737467011437, 0.483507137279958, 0.964194590225816, 0.424605934647843, 0.526403775205836, 0.0640464650932699, 0.0568529251031578, 0.356340519152582, 0.931965528521687, 0.193561173975468, 0.479296899167821, 0.0902100375387818, 0.193010108778253, 0.843470606952906, 0.522111324360594, 0.02260248712264, 0.0965414880774915, 0.218960575060919, 0.408842392731458, 0.509263135492802, 0.6524140634574, 0.886049021501094, 0.319913276238367, 0.684978582663462, 0.727717032423243, 0.799588667228818, 0.319833129178733, 0.53731218748726, 0.330802268814296, 0.0165774102788419, 0.996445463271812, 0.0122569559607655, 0.156451139133424, 0.77616128185764, 0.474843282951042, 0.53482626657933, 0.808870400302112, 0.0132711688056588, 0.0902288912329823, 0.381599060259759, 0.963587822858244, 0.443367179250345, 0.168650812003762, 0.831259431550279, 0.713551336433738, 0.471739704022184, 0.0867450651712716, 0.639163743937388, 0.117654716363177, 0.961665768176317, 0.365035569528118, 0.613855368457735, 0.219226294895634, 0.861120329005644, 0.553789504803717, 0.183472957462072, 0.400601633591577, 0.431715029291809, 0.49635273613967, 0.3401249863673, 0.604490486904979, 0.444052859675139, 0.877381344791502, 0.649120914982632, 0.948978761676699, 0.277390757808462, 0.269042202737182, 0.320678623858839, 0.53894773684442, 0.606517099076882, 0.239013975486159, 0.582317018415779, 0.944743593456224, 0.835120344068855, 0.0314605899620801, 0.694684075191617, 0.840202045161277, 0.765394650632516, 0.587697541806847, 0.97536082030274, 0.0567836835980415, 0.98782789031975, 0.106551403179765, 0.574677334632725, 0.492820403072983, 0.719826851971447, 0.351434181677178, 0.550989351468161, 0.892678990028799, 0.181117589119822, 0.992122275521979, 0.393051199847832, 0.0237275306135416, 0.138726542470977, 0.909890068229288, 0.527506328653544, 0.633417265489697, 0.533788215834647, 0.456373162567616, 0.461107414448634, 0.502114272676408, 0.431897916132584, 0.654781563207507, 0.579276182455942, 0.398556945845485, 0.0537518884520978, 0.666811214294285, 0.536185747012496, 0.0711149836424738, 0.717369053745642, 0.964436810696498, 0.0180597826838493, 0.592537095304579, 0.188259136164561, 0.267680424731225, 0.891949257813394, 0.682453810237348, 0.135640903376043, 0.728210260625929, 0.629771661479026, 0.606483788695186, 0.207709322450683, 0.306301606353372, 0.118695045122877, 0.872133964439854, 0.018711103592068, 0.872145284432918, 0.180902949068695, 0.141637519467622, 0.650260895257816, 0.828383869724348, 0.990253292722628, 0.775957483565435, 0.588681090855971, 0.734274527989328, 0.587393058929592, 0.845313932048157, 0.130331058055162, 0.0579860957805067, 0.939335173694417, 0.50166382198222, 0.278292810311541, 0.491765622515231, 0.964531478704885, 0.179410526761785), actual = c(1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.622916508000344, 0.439192387042567, 0.618800505064428, 0.527013565879315, 0.315067020244896, 0.237633937271312, 0.512325109448284, 0.287228869972751, 0.0825695805251598, 0.185676705557853, 0.21115060406737, 0.49177758465521, 0.120035595959052, 0.707809462677687, 0.07941006985493, 0.622954169288278, 0.768055564491078, 0.0328948234673589, 0.285778807243332, 0.0995802786201239, 0.797630849294364, 0.946723855799064, 0.802660133456811, 0.233616614714265, 0.463755875825882, 0.399444109294564, 0.844206175999716, 0.323736202437431, 0.992441289126873, 0.76679365709424, 0.0895581431686878, 0.958054542541504, 0.139776547439396, 0.0445377451833338, 0.531034250045195, 0.931155978702009, 0.300838608760387, 0.69590291660279, 0.915038388688117, 0.271733827656135, 0.0976726762019098, 0.482626069802791, 0.551145712379366, 0.180694484850392, 0.564833730459213, 0.262315351748839, 0.56484410376288, 0.784151695668697, 0.582466208608821, 0.969426865456626, 0.364737567957491, 0.517918935744092, 0.182256668340415, 0.171600909670815, 0.71951466659084, 0.682215532753617, 0.615445349365473, 0.848949543666095, 0.650858923094347, 0.579977074405178, 0.445523019880056, 0.497196892974898, 0.49911596882157, 0.0148007445968688, 0.738170090597123, 0.420058370567858, 0.488278163596988, 0.0147855319082737, 0.687780624488369, 0.912524210521951, 0.574614292010665, 0.504406907362863, 0.166348269209266, 0.708709827857092, 0.35553229926154, 0.542561690555885, 0.174611578928307, 0.0691225244663656, 0.386808954644948, 0.72296091215685, 0.267114603193477, 0.53538480005227, 0.294987743487582, 0.469603634672239, 0.8616335936822, 0.120731366798282, 0.428665394429117, 0.909537890926003, 0.0418295322451741, 0.273442550562322, 0.911269289907068, 0.686611713143066, 0.821252725785598, 0.56111522205174, 0.63115220121108, 0.145922103663906, 0.522181336767972, 0.408518011681736, 0.935206703841686, 0.942859481088817, 0.0843978442717344, 0.27591000450775, 0.207610720070079, 0.340573000255972, 0.498993555549532, 0.0676863479893655, 0.559404778759927, 0.547761495225132, 0.949765311088413, 0.232125895563513, 0.176073547452688, 0.442468348424882, 0.886809090152383, 0.446516183204949, 0.632563431747258, 0.401755179977044, 0.566074538510293, 0.729954476235434, 0.00413586548529565, 0.307240474736318, 0.791044515790418, 0.178742743097246, 0.160190667724237, 0.500581292901188, 0.625735183246434, 0.365581420483068, 0.79707799339667, 0.598909503314644, 0.628528742585331, 0.951059804065153, 0.277239418588579, 0.985771111445501, 0.936354158213362, 0.172535243444145, 0.108718477655202, 0.359126459341496, 0.559751970460638, 0.973123617935926, 0.721644905162975, 0.458523565204814, 0.112406058702618, 0.63379780203104, 0.684887662297115, 0.14754537306726, 0.0144930724054575, 0.563897937303409, 0.904868840472773, 0.0660556133370847, 0.837239287560806, 0.681372190592811, 0.766852021217346, 0.779073546407744, 0.00596883846446872, 0.203676665667444, 0.540766905294731, 0.347901098197326, 0.37202069349587, 0.849229025188833, 0.637081897817552, 0.284460443072021, 0.837736091110855, 0.0606447502505034, 0.0484519214369357, 0.465943200048059, 0.581017031334341, 0.552764614811167, 0.86156561691314, 0.289826102089137, 0.76794474478811, 0.631566845113412, 0.0994915044866502, 0.51084663788788, 0.889360333327204, 0.152533881599084, 0.0478464905172586, 0.911737591028214, 0.945565820205957, 0.626706762239337, 0.543459438486025, 0.327135720755905, 0.619072025176138, 0.865763371111825, 0.181876667542383, 0.326761181233451, 0.721631036372855, 0.953466365812346, 0.37818511785008, 0.12709320220165, 0.408677836647257, 0.202973772771657, 0.358628020389006, 0.15660739177838, 0.664059564005584, 0.634044666076079, 0.785666628275067, 0.187798243714496, 0.338857753202319, 0.736246205400676, 0.0671391985379159, 0.593428767286241, 0.209753630449995, 0.967439090134576, 0.581992308376357, 0.0238852351903915, 0.499717824859545, 0.841445019235834, 0.870058414293453, 0.78499815124087, 0.707714440533891, 0.503875873284414, 0.172130312304944, 0.477815896272659, 0.570578733924776, 0.24293162045069, 0.058418771950528, 0.85453060711734, 0.518266105093062, 0.139100033789873, 0.899135599611327, 0.472295515006408, 0.194890196900815, 0.0939991346094757, 0.261523308930919, 0.599745971616358, 0.903939301148057, 0.174782578134909, 0.426113537978381, 0.931999542051926, 0.446674927370623, 0.57044649659656, 0.922831372357905, 0.375018257880583, 0.294748279033229, 0.569056256907061, 0.49945185566321, 0.322352794697508, 0.689849044429138, 0.8708354474511, 0.651777634397149, 0.757528885966167, 0.713750074850395, 0.579063411103562, 0.606288849376142, 0.209402607521042, 0.556179202860221, 0.707546310033649, 0.612200807547197, 0.710329368710518, 0.523473310982808, 0.647797279292718, 0.488378464709967, 0.200311844469979, 0.575237880460918, 0.60153662157245, 0.0378082166425884, 0.424869093345478, 0.791670633712783, 0.603955632308498, 0.710233802907169, 0.879390743095428, 0.0239097976591438, 0.387572042178363, 0.584961482090876, 0.351229103049263, 0.633949204348028, 0.458978098118678, 0.541127864038572, 0.441653864691034, 0.724350575124845, 0.27127268887125, 0.48534016800113, 0.18216701829806, 0.259736839681864, 0.200039060320705, 0.983677367679775, 0.819802033016458, 0.771238304208964, 0.0773515703622252, 0.44328008685261, 0.662651563528925, 0.362929079914466, 0.129765205318108, 0.631435354240239, 0.873355572111905, 0.78053845022805, 0.0027970839291811, 0.921883760485798, 0.601928534917533, 0.730999199440703, 0.382264708867297, 0.0603500830475241, 0.431291142711416, 0.403727720258757, 0.582901007728651, 0.344736794475466, 0.961752188624814, 0.459277217974886, 0.423803647281602, 0.926024755230173, 0.0870010682847351, 0.537474655546248, 0.756407975452021, 0.421767532592639, 0.433787552872673, 0.371634848183021, 0.176678339485079, 0.935474025085568, 0.871882324805483, 0.486440130276605, 0.845588860101998, 0.11104768468067, 0.537268328946084, 0.5658920397982, 0.493601513793692, 0.76526496373117, 0.82438550144434, 0.748751320177689, 0.127800352871418, 0.469172231154516, 0.211544658290222, 0.816987388534471, 0.579533224925399, 0.312313687754795, 0.828570319106802, 0.901913619367406, 0.66655244701542, 0.201733058551326, 0.166065457509831, 0.727484058588743, 0.100005235522985, 0.460299793398008, 0.454124648589641, 0.186063247034326, 0.4447803767398, 0.225488814758137, 0.358902551000938, 0.740163716953248, 0.746322013437748, 0.176175845088437, 0.850675905356184, 0.994518164079636, 0.812880465993658, 0.909059295896441, 0.442612013313919, 0.878789930138737, 0.388681111158803, 0.800491257337853, 0.689151409082115, 0.986978320172057, 0.81141570280306, 0.245149745140225, 0.166575383394957, 0.826594985788688, 0.108049234142527, 0.916706233751029, 0.746579118771479, 0.986705844523385, 0.347803301177919, 0.480397883336991, 0.101398134604096, 0.215533647220582, 0.619459837209433, 0.922475394327193, 0.0938548371195793, 0.442403945140541, 0.155539185507223, 0.600983719108626, 0.362205025739968, 0.501148154260591, 0.0442326371558011, 0.659256079234183, 0.170441519469023, 0.86119702225551, 0.570454719010741, 0.183858118485659, 0.788230945821851, 0.788367175264284, 0.0296645504422486, 0.289065417833626, 0.668229688424617, 0.565668388735503, 0.82234539068304, 0.0318689504638314, 0.532378383679315, 0.798705804860219, 0.242791792843491, 0.792497257934883, 0.352619883604348, 0.64184340252541, 0.275958176702261, 0.974278435576707, 0.892339224927127, 0.403865238418803, 0.962955552618951, 0.322649132693186, 0.933561434969306, 0.597788043087348, 0.820325495209545, 0.668029139982536, 0.38538917619735, 0.80361513630487, 0.221248732181266, 0.663187932223082, 0.633958668448031, 0.879978730110452, 0.73004958499223, 0.853105150628835, 0.0365707136224955, 0.390682784840465, 0.373699968447909, 0.946920674294233, 0.0649197322782129, 0.176102936966345, 0.820422326447442, 0.741104666143656, 0.363225381821394, 0.463629233185202, 0.494107746053487, 0.688273148145527, 0.112856391118839, 0.466736383503303, 0.305838060099632, 0.267072466667742, 0.244152281433344, 0.846921565942466, 0.80799545510672, 0.636857198551297, 0.7234150134027, 0.140117176575586, 0.157418102258816, 0.56572315050289, 0.253337011439726, 0.591755987610668, 0.232192493742332, 0.592303003184497, 0.226711537688971, 0.296375352423638, 0.71568070910871, 0.247610547114164, 0.911973978858441, 0.199240928981453, 0.787567773135379, 0.736782974097878, 0.662211765302345, 0.725758504355326, 0.49959154939279, 0.416676323860884, 0.445421069860458, 0.0368600408546627, 0.260368624469265, 0.657660099677742, 0.152300318935886, 0.458421006565914, 0.306918775429949, 0.439501037355512, 0.100754947867244, 0.654038627166301, 0.817077894229442, 0.54537677182816, 0.973623802652583, 0.281144480919465, 0.665710524655879, 0.0749244028702378, 0.10881891194731, 0.411733641754836, 0.601814249996096, 0.639897385844961, 0.244339835597202, 0.598240749677643, 0.342234000563622, 0.520444999216124, 0.76384927961044, 0.580527646234259, 0.239512688247487, 0.291910015745088, 0.990407075732946, 0.433496230281889, 0.379673842573538, 0.142688527004793, 0.675131170544773, 0.204485289985314, 0.492402998963371, 0.0853875258471817, 0.729563089786097, 0.676553873810917, 0.344716313295066, 0.96802143799141, 0.484925109660253, 0.0967223572079092, 0.802680717082694, 0.803729784442112, 0.344759649829939, 0.44751703646034, 0.0467821257188916, 0.253175107995048, 0.48083366593346, 0.993198808981106, 0.403347220504656, 0.519361626822501, 0.97950470354408, 0.834462201455608, 0.914826235035434, 0.60394129389897, 0.253985468531027, 0.95606911345385, 0.0653923705685884, 0.772766462294385, 0.814664193894714, 0.40638576564379, 0.628453941084445, 0.125381577527151, 0.683874452486634, 0.162829048698768, 0.169552974868566, 0.185016084229574, 0.397592311026528, 0.714305771980435, 0.38040612381883, 0.0418009844142944, 0.557027315022424, 0.439120061695576, 0.897246874403208, 0.312815620796755, 0.973267003428191, 0.290245894342661, 0.502960304031149, 0.0915369524154812, 0.677197414683178, 0.298934019170702, 0.179507818771526, 0.841566996648908, 0.720147385029122, 0.385462522041053, 0.919983529485762, 0.637884672731161, 0.431385717121884, 0.526487551629543, 0.862601812696084, 0.459393656579778, 0.818619603523985, 0.658037955872715, 0.0471309740096331, 0.592665640870109, 0.538490584352985, 0.369847478577867, 0.306612175889313, 0.112573251361027, 0.584781196434051, 0.322885705623776, 0.705405890475959, 0.494374304544181, 0.172935905866325, 0.165326488669962, 0.149364225799218, 0.830621533328667, 0.659635151503608, 0.847549255006015, 0.223689032252878, 0.304461050080135, 0.427155867684633, 0.500026143621653, 0.170871440554038, 0.935962322400883, 0.674221682362258, 0.692273400025442, 0.0677503519691527, 0.460028230445459, 0.874566247686744, 0.619057942880318, 0.654488540487364, 0.331040415680036, 0.431224527303129, 0.727298165438697, 0.920934410532936, 0.238288194639608, 0.260870952391997, 0.269686108687893, 0.9144653796684, 0.56244048057124, 0.357254573144019, 0.132411939557642, 0.691954679321498, 0.477453615749255, 0.163773274281994, 0.820270115043968, 0.767690852982923, 0.162465984001756, 0.918207414448261, 0.532726334407926, 0.0129905119538307, 0.592051309067756, 0.0296603217720985, 0.908387212315574, 0.398036628495902, 0.545984183438122, 0.867148456629366, 0.790150835644454, 0.399820588994771, 0.536137801362202, 0.696092621656135, 0.600465782918036, 0.352657651063055, 0.531896040309221, 0.883580443682149, 0.622957278974354, 0.531938954954967, 0.413846004288644, 0.843888861360028, 0.520262910053134, 0.446524772560224, 0.0716809625737369, 0.201340486062691, 0.0134977402631193, 0.949843291658908, 0.155643139500171, 0.640223854919896, 0.304463137406856, 0.327639386290684, 0.728021922754124, 0.487279294058681, 0.140868441900238, 0.387704954482615, 0.393671602709219, 0.716665612766519, 0.56927228718996, 0.284329892368987, 0.659873084630817, 0.497024744283408, 0.0791152631863952, 0.425809170817956, 0.479846313362941, 0.847229921724647, 0.0634554498828948, 0.26711961789988, 0.149530295515433, 0.444741254206747, 0.603071396471933, 0.982288572704419, 0.18879916286096, 0.798126728506759, 0.904870895436034, 0.1549313578289, 0.809416653588414, 0.644857133273035, 0.499453322729096, 0.383231823332608, 0.919569323305041, 0.870558450231329, 0.766171808587387, 0.320867822738364, 0.283458996796981, 0.564381322357804, 0.957657982362434, 0.81649687839672, 0.703467849874869, 0.444429799681529, 0.602396509610116, 0.168832902330905, 0.200990845216438, 0.00404140772297978, 0.253185671987012, 0.707426643930376, 0.267696570139378, 0.506795522756875, 0.714813855011016, 0.0907261676620692, 0.697590165771544, 0.889092811848968, 0.623840325046331, 0.937585674691945, 0.500038285972551, 0.477129301754758, 0.110423804959282, 0.512258055387065, 0.0478334855288267, 0.166159273125231, 0.325348446844146, 0.360102005768567, 0.409727776190266, 0.995144866406918, 0.402219698298723, 0.979337461059913, 0.102846664376557, 0.115638725226745, 0.904359788633883, 0.66063602338545, 0.727573192911223, 0.049366437131539, 0.870778178330511, 0.29957015812397, 0.9684166200459, 0.140280951512977, 0.0985811031423509, 0.648735058959574, 0.386121949646622, 0.933080229442567, 0.489491343032569, 0.762773590628058, 0.458921768702567, 0.807914714561775, 0.706759946420789, 0.821900086011738, 0.431366313947365, 0.681297598173842, 0.783597239293158, 0.838860124582425, 0.727553551783785, 0.908116984646767, 0.0121195011306554, 0.303493180777878, 0.259974143933505, 0.960245538735762, 0.789583151461557, 0.769173009088263, 0.229550394229591, 0.631866740295663, 0.125413165660575, 0.701010838616639, 0.176788204582408, 0.246850098250434, 0.821018624119461, 0.0039452351629734, 0.739931562216952, 0.480322869960219, 0.497193914139643, 0.159924457781017, 0.398398992372677, 0.22385817905888, 0.912030479870737, 0.0265954488422722, 0.690954691031948, 0.0439100093208253, 0.689823069842532, 0.292642432730645, 0.571692704223096, 0.356409889878705, 0.543307768646628, 0.516105866292492, 0.444658859632909, 0.951157737756148, 0.922032467555255, 0.338966065552086, 0.893918910995126, 0.973246978363022, 0.0220900613348931, 0.159930997062474, 0.0325369162019342, 0.919771062908694, 0.622118227882311, 0.264523120597005, 0.839352349983528, 0.392148327780887, 0.460430229315534, 0.934820035239682, 0.836287793470547, 0.815268059261143, 0.0884592786896974, 0.40158040355891, 0.932595364982262, 0.178243834292516, 0.225464349845424, 0.69237817870453, 0.348956481087953, 0.605272300774232, 0.323685315437615, 0.314221416600049, 0.0934515919070691, 0.807473949622363, 0.297129106940702, 0.71678562508896, 0.748938901582733, 0.257394238840789, 0.386313701514155, 0.874736876226962, 0.154231993947178, 0.600404870230705, 0.675129957264289, 0.853116757934913, 0.289856065995991, 0.264643292408437, 0.749830146087334, 0.910453835502267, 0.444200293160975, 0.904557680012658, 0.103255243040621, 0.357897132169455, 0.946792004397139, 0.38331963843666, 0.976956286234781, 0.854465746786445, 0.970561488997191, 0.611685574520379, 0.23973699240014, 0.761714484076947, 0.456316334195435, 0.955140913836658, 0.31566343200393, 0.597602599533275, 0.841206396231428, 0.882790033239871, 0.619909406406805, 0.329799056984484, 0.931731471326202, 0.506590154953301, 0.290958185680211, 0.0451081383507699, 0.800105811329558, 0.653873632196337, 0.820218190317973, 0.489856736734509, 0.297523080604151, 0.680644522886723, 0.462795250350609, 0.691264757886529, 0.614547623554245, 0.193616501288489, 0.0131872626952827, 0.023372745141387, 0.684867235366255, 0.384788064286113, 0.776811642805114, 0.317070655990392, 0.275126541033387, 0.622717915102839, 0.997450492577627, 0.703423629747704, 0.57378243538551, 0.708631522022188, 0.573531416477636, 0.927510483656079, 0.379361533792689, 0.0351861198432744, 0.636931452667341, 0.244531311327592, 0.471300696022809, 0.115210993913934, 0.249427856644616, 0.507559010991827, 0.0238409279845655, 0.858957263408229, 0.318535365397111, 0.745308400830254, 0.348619798664004, 0.0213081520050764, 0.206440845737234, 0.786570645635948, 0.894176084082574, 0.925260449294001, 0.968548444798216, 0.0371826642658561, 0.187626800732687, 0.561875458341092, 0.446494156261906, 0.424284873064607, 0.281916607869789, 0.545376837253571, 0.0997387855313718, 0.401150942081586, 0.34527772013098, 0.562749990262091, 0.311082279542461, 0.982337852939963, 0.342171739088371, 0.464160140836611, 0.730526351137087, 0.734281950863078, 0.977172823622823, 0.785095505882055, 0.874474050477147, 0.983987641753629, 0.409154987661168, 0.555028755683452, 0.106644722400233, 0.691873470321298, 0.0994211083743721, 0.404699456878006, 0.35032547893934, 0.987408084562048, 0.0396128341089934, 0.535979484207928, 0.416402701055631, 0.525015087798238, 0.366805847035721, 0.806851407978684, 0.61658904934302, 0.329556269804016, 0.823474829085171, 0.924729969585314, 0.0973308007232845, 0.906848747516051, 0.291802180930972, 0.770801840815693, 0.802010672166944, 0.888252970995381, 0.784015850396827, 0.849237997084856, 0.706868870882317, 0.865081646479666, 0.255902090109885, 0.657721946947277, 0.842313517350703, 0.528442273614928, 0.948128553340212, 0.653995675733313, 0.803873646073043, 0.627253229729831, 0.467563468264416, 0.215349740115926, 0.0172918650787324, 0.109405598603189, 0.787539605749771, 0.157680639065802, 0.690567956771702, 0.122388140764087, 0.803765915334225, 0.421038918197155, 0.3630308189895, 0.808773993514478, 0.999351739883423, 0.383348958566785, 0.835917675867677, 0.478512894362211, 0.757970117265359, 0.0808608252555132, 0.769676466239616, 0.707921489374712, 0.382860914804041, 0.24040217557922, 0.0739130892325193, 0.201313221827149, 0.0298139976803213, 0.846185798989609, 0.475887339329347, 0.119099596980959, 0.906682461732998, 0.595935895340517, 0.490352927474305, 0.717623794451356, 0.364050720119849, 0.246909168781713, 0.316532524535432, 0.457066375995055, 0.171011564088985, 0.300531738437712, 0.173947732895613, 0.704641731921583, 0.670515067642555, 0.413081882754341, 0.788988950196654, 0.299560675863177, 0.84208873193711, 0.749001792166382, 0.102683874079958, 0.792363794287667, 0.843935189535841, 0.270260576158762, 0.0973754876758903, 0.53081501298584, 0.144676625961438, 0.361477451631799, 0.295162731315941, 0.50850235321559, 0.677565150428563, 0.983731651445851, 0.353281994583085, 0.0605490307789296, 0.519000452943146, 0.0680744436103851, 0.623733147745952, 0.805481409886852, 0.696373631246388, 0.628180101746693, 0.799252049764618, 0.270765696186572, 0.500489328522235, 0.485120788915083, 0.503355772234499, 0.356103223515674, 0.0037050349637866, 0.574045865330845, 0.643337275832891, 0.35069429106079, 0.495348798343912, 0.160654510837048, 0.594945991411805, 0.470113360555843, 0.342527193250135, 0.920538170728832, 0.775793302804232, 0.0489917290396988, 0.294254655949771, 0.855891623999923, 0.905080230440944, 0.912622366566211, 0.743231044849381, 0.943337101954967, 0.0538961819838732, 0.173291775630787, 0.683858711505309, 0.489573300816119), actual = c(0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"