Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4351967        0.4351967 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4351967        0.4487368        0.5105592 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.0682041896507144, 0.140444246353582, 0.677762484177947, 0.828753649955615, 0.349695071112365, 0.139651008183137, 0.663374191150069, 0.239301207941025, 0.322255458217114, 0.860488718142733, 0.602166065014899, 0.509765934431925, 0.941772245801985, 0.832978407852352, 0.771858244203031, 0.224853259976953, 0.154621700290591, 0.882150492630899, 0.815439626341686, 0.458442606264725, 0.653056053910404, 0.118807408725843, 0.202335260808468, 0.87078538024798, 0.022400040877983, 0.659103934187442, 0.915045312838629, 0.285906004253775, 0.934574937215075, 0.940076224971563, 0.758455093484372, 0.673115731449798, 0.832397509366274, 0.509925515856594, 0.218458669027314, 0.776649964274839, 0.184580717934296, 0.463848019484431, 0.166521419305354, 0.878630952676758, 0.994385150494054, 0.560949106235057, 0.819065080722794, 0.685603667981923, 0.222103491891176, 0.29306912003085, 0.0317286644130945, 0.21736603602767, 0.103792768903077, 0.0680389790795743, 0.591614483157173, 0.732795560965315, 0.461323460331187, 0.47188944532536, 0.861390433041379, 0.369283841690049, 0.793126397766173, 0.818902905099094, 0.821516151772812, 0.329381513642147, 0.463606036035344, 0.210176329128444, 0.0460026548244059, 0.105763016268611, 0.943038090365008, 0.0321770205628127, 0.503480202285573, 0.925742415245622, 0.095692295813933, 0.958478559507057, 0.588408244075254, 0.6582051650621, 0.37484601023607, 0.775283264694735, 0.194366663461551, 0.400612743105739, 0.753794451942667, 0.590199094731361, 0.46925184619613, 0.399473698111251, 0.0509186417330056, 0.216245845193043, 0.874687071191147, 0.148143007885665, 0.116113962605596, 0.932556367479265, 0.792872838443145, 0.922105316538364, 0.547571291215718, 0.510868723504245, 0.835532212164253, 0.794549310812727, 0.249845334794372, 0.844549949048087, 0.294600939843804, 0.138092713896185, 0.0906489677727222, 0.350327558349818, 0.447723455028608, 0.754558945074677, 0.501074952306226, 0.689213084522635, 0.219526368193328, 0.833022449864075, 0.875108037609607, 0.0667307279072702, 0.121452998602763, 0.555059536127374, 0.976343153975904, 0.780831372831017, 0.734694015467539, 0.50675959396176, 0.537227204302326, 0.722131754504517, 0.674772135447711, 0.333875183714554, 0.155847376212478, 0.512337944470346, 0.804265609011054, 0.436963590560481, 0.0943827831652015, 0.738680147565901, 0.946378876687959, 0.993793775560334, 0.488553416915238, 0.698315055808052, 0.0635696148965508, 0.223188124829903, 0.332950031384826, 0.0519965363200754, 0.159329874208197, 0.811066779308021, 0.881640765350312, 0.769629673333839, 0.470018068794161, 0.22893614275381, 0.265768781770021, 0.0163171836175025, 0.738259048201144, 0.162379717687145, 0.468191909138113, 0.591954427072778, 0.0949449597392231, 0.94648903934285, 0.0368839388247579, 0.458878765581176, 0.475996374851093, 0.272472389508039, 0.26955151068978, 0.339250100776553, 0.288163207704201, 0.878043519100174, 0.489006336778402, 0.315146899549291, 0.941045543178916, 0.077285606879741, 0.0436396624427289, 0.399694266496226, 0.690714115742594, 0.441720539238304, 0.406224314356223, 0.241313251899555, 0.279317816719413, 0.0447627585381269, 0.533793560229242, 0.0153928706422448, 0.532043750630692, 0.565992300398648, 0.843694896204397, 0.261127630248666, 0.374898650916293, 0.271421687211841, 0.680633089970797, 0.812407493591309, 0.154105931287631, 0.454890195978805, 0.400320186279714, 0.53793302224949, 0.543298140633851, 0.0242175413295627, 0.440161554142833, 0.647116990759969, 0.559081350453198, 0.467507081804797, 0.821967337280512, 0.351061330409721, 0.774221380241215, 0.124975580489263, 0.177415092941374, 0.0792978776153177, 0.751250182744116, 0.879647476132959, 0.911143695237115, 0.0357691207900643, 0.652865725802258, 0.621795864542946, 0.432154740672559, 0.817075769416988, 0.565342966932803, 0.184822911629453, 0.222490086685866, 0.387173436814919, 0.778008835623041, 0.312461073277518, 0.0933380501810461, 0.0355198262259364, 0.577277990290895, 0.273878459585831, 0.892506445525214, 0.795090480474755, 0.7452574565541, 0.295859788311645, 0.357908905716613, 0.665352446958423, 0.225005573360249, 0.175176522461697, 0.837959714233875, 0.795222736429423, 0.188871300779283, 0.263207467505708, 0.827409357530996, 0.365698812995106, 0.499773637624457, 0.537848739884794, 0.21056771860458, 0.540766240097582, 0.770248094340786, 0.457569440128282, 0.301664551487193, 0.862272094236687, 0.858184016542509, 0.543546152999625, 0.865523794665933, 0.0905979098752141, 0.873180884402245, 0.13735658954829, 0.903146204072982, 0.934541959082708, 0.899570876732469, 0.816460534464568, 0.284232194768265, 0.49000430922024, 0.263875217176974, 0.975207039853558, 0.904495344497263, 0.50547291059047, 0.579587371554226, 0.417384550441056, 0.0634102462790906, 0.576125821331516, 0.429634387372062, 0.801068807020783, 0.560602768091485, 0.502914036856964, 0.727983963442966, 0.586630876641721, 0.462634549476206, 0.966782455099747, 0.0397866326384246, 0.0207196574192494, 0.321955702500418, 0.33889427757822, 0.331594537012279, 0.147377448156476, 0.201107257744297, 0.621277343248948, 0.388374086003751, 0.452605441212654, 0.502340614795685, 0.108432976994663, 0.928109898231924, 0.823452283861116, 0.893568749306723, 0.308225558605045, 0.461488595232368, 0.151790556730703, 0.614327885676175, 0.894868204602972, 0.120411269599572, 0.305148623883724, 0.89301325683482, 0.317083925474435, 0.641449822112918, 0.948792462004349, 0.406340804882348, 0.829100298229605, 0.439978634472936, 0.0125572788529098, 0.374558806652203, 0.504729094449431, 0.126895443769172, 0.245247472543269, 0.175849091494456, 0.350590783171356, 0.850052301073447, 0.487922043772414, 0.790682165650651, 0.288596165599301, 0.58766606147401, 0.78264846582897, 0.841048349859193, 0.182789411395788, 0.368763537378982, 0.681069125654176, 0.959483472397551, 0.807243102928624, 0.546070114476606, 0.958455577492714, 0.402437175158411, 0.518684432841837, 0.478812005138025, 0.801965435268357, 0.712419836549088, 0.279282304691151, 0.165588871808723, 0.738604198442772, 0.792155558941886, 0.931086858501658, 0.242127637146041, 0.18431887589395, 0.392405123217031, 0.0482884547673166, 0.00469190534204245, 0.558280504308641, 0.28617018321529, 0.314003996085376, 0.814651373773813, 0.226294180145487, 0.904052847530693, 0.715179080842063, 0.73165796790272, 0.204551605973393, 0.554311782354489, 0.119872342329472, 0.885213290574029, 0.221288283588365, 0.963436458492652, 0.659409959800541, 0.606458558700979, 0.161889845272526, 0.769441593671218, 0.460001516854391, 0.0541864784900099, 0.669826291967183, 0.221717628417537, 0.279837009031326, 0.197049866197631, 0.585433836327866, 0.509607308311388, 0.744369195541367, 0.305938204750419, 0.167580371024087, 0.863808672176674, 0.574805117212236, 0.267118523828685, 0.289555663708597, 0.130834252107888, 0.526546803768724, 0.988342543365434, 0.904408579226583, 0.49806417664513, 0.302356455242261, 0.522143980953842, 0.03870501043275, 0.176605649059638, 0.428235130151734, 0.859925107331946, 0.281896320870146, 0.309326728573069, 0.637311689788476, 0.261569608468562, 0.506824281997979, 0.13569989730604, 0.486168753821403, 0.192560816649348, 0.765630948357284, 0.756166392704472, 0.722050575306639, 0.77045320882462, 0.881263325922191, 0.538582249777392, 0.372731544775888, 0.697176209185272, 0.419429639121518, 0.907199064269662, 0.352252987679094, 0.285866419551894, 0.80508812260814, 0.990179282613099, 0.820315667660907, 0.902028900105506, 0.001048389589414, 0.972223307471722, 0.731870677787811, 0.620974894613028, 0.659279425395653, 0.803987855324522, 0.574402185156941, 0.757909546606243, 0.773899367311969, 0.123769288416952, 0.606089615961537, 0.711313784820959, 0.882938398281112, 0.486767327645794, 0.504256718093529, 0.492234945297241, 0.537222535116598, 0.521774316905066, 0.906141394050792, 0.431551782647148, 0.271232966566458, 0.541006693383679, 0.372499791672453, 0.425094904145226, 0.308372272877023, 0.764174959855154, 0.493513219058514, 0.405320185469463, 0.911918416619301, 0.886656164191663, 0.453288493677974, 0.235591988312081, 0.595176694681868, 0.229417052818462, 0.962644457118586, 0.411784930853173, 0.367786392802373, 0.191211245255545, 0.387100273510441, 0.383792259730399, 0.204506380250677, 0.333103612530977, 0.670584337553009, 0.174065637867898, 0.121478436049074, 0.0722571599762887, 0.0701858124230057, 0.173062424408272, 0.548914131708443, 0.401122120441869, 0.757044864818454, 0.881970550632104, 0.779642052017152, 0.211930776480585, 0.43310943688266, 0.459830993553624, 0.288885666290298, 0.963880940107629, 0.0710645674262196, 0.301642586709931, 0.344527677400038, 0.478759520687163, 0.541349709266797, 0.483706504106522, 0.991649771807715, 0.581493293866515, 0.454715368803591, 0.139044819399714, 0.0953417925629765, 0.605808291351423, 0.444179634563625, 0.849149293731898, 0.460872606607154, 0.962227291660383, 0.429037519032136, 0.15179376443848, 0.769768494181335, 0.904790858738124, 0.138228378258646, 0.790334091288969, 0.326552717247978, 0.901056190021336, 0.780918495496735, 0.943653172813356, 0.0266286465339363, 0.339613430434838, 0.26659017521888, 0.882057308219373, 0.939318198012188, 0.153922378784046, 0.600760583067313, 0.0933544298168272, 0.074259425047785, 0.295721541857347, 0.168041283497587, 0.349312745966017, 0.756394508527592, 0.0351758047472686, 0.838442360516638, 0.992136993678287, 0.534465171163902, 0.113642779411748, 0.994882092811167, 0.719122804002836, 0.563919199630618, 0.00386277842335403, 0.600398103473708, 0.154388049151748, 0.258896089391783, 0.184157453943044, 0.654864069540054, 0.8744396653492, 0.822719887830317, 0.0418119125533849, 0.120942752342671, 0.973045795457438, 0.595550378784537, 0.916645204415545, 0.253749146126211, 0.43163405964151, 0.0341423361096531, 0.407936765346676, 0.96087235910818, 0.60892671556212, 0.789877786301076, 0.923344394890592, 0.118401929736137, 0.0470281601883471, 0.6794727682136, 0.424756442429498, 0.724653450772166, 0.187016424024478, 0.142415896058083, 0.806671735597774, 0.58664354425855, 0.851216458715498, 0.346847401699051, 0.0497630273457617, 0.587529812473804, 0.737651995616034, 0.524698713794351, 0.157483025919646, 0.376312548993155, 0.312377258436754, 0.960519638843834, 0.1301372081507, 0.530335700605065, 0.827294406713918, 0.771464197896421, 0.402341924142092, 0.316412708023563, 0.115181316388771, 0.883888326352462, 0.266411914490163, 0.466219889931381, 0.187987328507006, 0.076486092293635, 0.509071968961507, 0.597724966937676, 0.93783430987969, 0.413490748265758, 0.321873279754072, 0.936595515348017, 0.891000867588446, 0.356897267280146, 0.25373974442482, 0.384114419110119, 0.737400630721822, 0.175127574009821, 0.744180460460484, 0.475840558996424, 0.253662263741717, 0.926624067360535, 0.961463934974745, 0.838577563874424, 0.650514597306028, 0.329634122783318, 0.246098623611033, 0.721182513050735, 0.836222189478576, 0.621211777208373, 0.314553258009255, 0.509193378733471, 0.821382261347026, 0.173451422248036, 0.301883135689422, 0.924565424909815, 0.9335122697521, 0.449005073169246, 0.860502568772063, 0.287630428327248, 0.778586659347638, 0.735120433848351, 0.490509049268439, 0.544977164128795, 0.000197765417397022, 0.262974687851965, 0.884729488287121, 0.299770475132391, 0.571592199848965, 0.915701986290514, 0.406220172764733, 0.0803428534418344, 0.638435460394248, 0.989944090135396, 0.42617366486229, 0.728255878435448, 0.0681927935220301, 0.791551084956154, 0.761874480172992, 0.492353053065017, 0.275196178117767, 0.175245739985257, 0.266603417461738, 0.426240868400782, 0.789292804896832, 0.473007716471329, 0.841664395760745, 0.91146552702412, 0.0193139479961246, 0.417138877790421, 0.186839613830671, 0.63586621847935, 0.116988116875291, 0.337970541091636, 0.820583486696705, 0.879932268988341, 0.287217525299639, 0.469349837396294, 0.778394956840202, 0.298347825184464, 0.963998220395297, 0.774488647468388, 0.333428774029016, 0.088877402478829, 0.836542035685852, 0.213822794146836, 0.837695349007845, 0.157393492991105, 0.936579237924889, 0.830090250587091, 0.838111105374992, 0.0859993915073574, 0.714404333848506, 0.813095576828346, 0.970942320534959, 0.405673212371767, 0.727372552501038, 0.232101607136428, 0.584078614832833, 0.239952841307968, 0.51248187944293, 0.139584527118132, 0.571826811414212, 0.311871679499745, 0.808947233483195, 0.0239343838766217, 0.311926160473377, 0.259502784116194, 0.844406329095364, 0.794344300404191, 0.907133015803993, 0.950733200181276, 0.623382757417858, 0.619472132530063, 0.309829394798726, 0.780197323299944, 0.237206998979673, 0.463232865324244, 0.388631392270327, 0.902568599674851, 0.266854798886925, 0.134992608334869, 0.482757257996127, 0.442042001057416, 0.226212215842679, 0.193315311102197, 0.984989307587966, 0.611873621819541, 0.581536906538531, 0.429955050116405, 0.663586388574913, 0.875095403054729, 0.45814598002471, 0.531132093863562, 0.163773040520027, 0.403174753068015, 0.499647620599717, 0.280697711277753, 0.749376063002273, 0.517973729642108, 0.676813416182995, 0.662755434401333, 0.62093804567121, 0.0399897315073758, 0.999528373591602, 0.353398317238316, 0.0125939939171076, 0.745178650831804, 0.0861914795823395, 0.444214895600453, 0.535080740228295, 0.7070868590381, 0.428199033252895, 0.533417667495087, 0.673642781097442, 0.84763922682032, 0.741962043568492, 0.900180659955367, 0.84641709481366, 0.355644553201273, 0.614709475776181, 0.131753898458555, 0.00887512438930571, 0.708553393604234, 0.531823125202209, 0.6634921960067, 0.151419463101774, 0.166548151057214, 0.195533910766244, 0.585564319975674, 0.180540146771818, 0.335255325771868, 0.0528048789128661, 0.438362045446411, 0.214309959439561, 0.751682767644525, 0.221847375622019, 0.0804570969194174, 0.779017567634583, 0.455585475545377, 0.914923619478941, 0.0839712379965931, 0.923637001076713, 0.955829591024667, 0.555184687487781, 0.747414752375335, 0.0482772598043084, 0.78441963205114, 0.218920168932527, 0.118478448363021, 0.806457936530933, 0.836649070028216, 0.3790873226244, 0.476509518455714, 0.773395945085213, 0.989216935355216, 0.698202220257372, 0.781427642796189, 0.580765353748575, 0.88737083482556, 0.427637940272689, 0.947200876427814, 0.123520786873996, 0.996517865918577, 0.0472695943899453, 0.818688641767949, 0.658563592936844, 0.313554019667208, 0.718195748981088, 0.318258890416473, 0.987171693705022, 0.971053464571014, 0.849892862839624, 0.264017623849213, 0.630652864929289, 0.0196878586430103, 0.336256633512676, 0.00711054890416563, 0.198809846770018, 0.897878043819219, 0.428664063336328, 0.303704731399193, 0.0155185461044312, 0.131489910418168, 0.917520842514932, 0.0590993240475655, 0.164983638795093, 0.847832279745489, 0.336994483834133, 0.381884743692353, 0.0189514835365117, 0.739845460979268, 0.396721125813201, 0.717114643892273, 0.128565194783732, 0.213472641538829, 0.381397492717952, 0.675633327802643, 0.19359565153718, 0.657002817606553, 0.939941341523081, 0.314435607753694, 0.881691447226331, 0.87171709863469, 0.824112114263698, 0.755319064715877, 0.812846964690834, 0.411345884436741, 0.699003834044561, 0.919286979129538, 0.304860771400854, 0.483115137554705, 0.906198606593534, 0.511882986407727, 0.625594175653532, 0.530343733960763, 0.729725695680827, 0.90159049583599, 0.942569938953966, 0.0489191815722734, 0.220264912117273, 0.469797309488058, 0.441511171637103, 0.671980999875814, 0.375508171506226, 0.575068827020004, 0.549711274914443, 0.386457356624305, 0.54839387582615, 0.232980537693948, 0.91493415646255, 0.304473639233038, 0.0296091679483652, 0.624495144933462, 0.472611419158056, 0.425632010912523, 0.904135846765712, 0.0234065032564104, 0.298130624461919, 0.338635172927752, 0.249781082617119, 0.399339476600289, 0.711897828383371, 0.531869913451374, 0.229843816021457, 0.0327609495725483, 0.609760851832107, 0.348858073819429, 0.352664084173739, 0.698984490940347, 0.555870810057968, 0.852684304350987, 0.917309041833505, 0.361129160504788, 0.508894686587155, 0.672155946725979, 0.0767336271237582, 0.857770718401298, 0.417565156472847, 0.330539075192064, 0.675811189226806, 0.93826227565296, 0.24482572125271, 0.639834355795756, 0.909717836184427, 0.449302074266598, 0.783453247277066, 0.726022328482941, 0.907629109220579, 0.837840995751321, 0.595636846264824, 0.098825843539089, 0.437459189910442, 0.0612259823828936, 0.161557022482157, 0.429880054667592, 0.827308924635872, 0.0483932439237833, 0.567035694839433, 0.144782309187576, 0.915800479473546, 0.444412126205862, 0.976670269854367, 0.360063720727339, 0.911894644843414, 0.672173895640299, 0.363193451426923, 0.547216905280948, 0.757694030413404, 0.483864677371457, 0.299622139427811, 0.130400562193245, 0.627034195233136, 0.261437785346061, 0.251406648894772, 0.000802946509793401, 0.772737559163943, 0.719951582839713, 0.157800578745082, 0.201652472140267, 0.00221101613715291, 0.602475709747523, 0.160603866679594, 0.377973814029247, 0.356323086889461, 0.386490234639496, 0.366027594311163, 0.370662863366306, 0.629752804292366, 0.0449227013159543, 0.535657310159877, 0.665056761819869, 0.8826463453006, 0.427792478585616, 0.174451701110229, 0.618262874893844, 0.537294043460861, 0.103781759971753, 0.670189381344244, 0.690027892589569, 0.931698261061683, 0.0092600726056844, 0.312710033729672, 0.480243853060529, 0.544129765359685, 0.234172157943249, 0.736488866154104, 0.36702746571973, 0.179337782086805, 0.149412896484137, 0.214071035850793, 0.0117941910866648, 0.639221185818315, 0.614056489197537, 0.037697859108448, 0.607837873278186, 0.819137684069574, 0.976342652691528, 0.0818183973897249, 0.581384824821725, 0.708312690025195, 0.245029928628355, 0.568166845478117, 0.591271988349035, 0.889272498665377, 0.390648158499971, 0.12141350377351, 0.236753077944741, 0.212178061716259, 0.637428294401616, 0.333295114804059, 0.947696237126365, 0.0668852517846972, 0.947536747669801, 0.3275662495289, 0.175929130753502, 0.380453854799271, 0.823832113761455, 0.927561882184818, 0.905674600973725, 0.684962900355458, 0.533971201395616, 0.483885924564674, 0.966231424827129, 0.704072602558881, 0.219430268974975, 0.829122794093564, 0.828493677312508, 0.49827334145084, 0.0216877225320786, 0.231612200615928, 0.948777144541964, 0.279196516377851, 0.71112188254483, 0.414202312240377, 0.677801923127845, 0.986615451518446, 0.920103064738214, 0.554893856868148, 0.508780782809481, 0.0804375745356083, 0.153861091239378, 0.572487564757466, 0.556971991201863, 0.111976097337902, 0.67846532841213, 0.114002058049664, 0.595880729611963, 0.569538405630738, 0.678751007653773, 0.858977433526888, 0.742923013400286, 0.77539035724476, 0.439691726583987, 0.653709391364828, 0.645531282527372, 0.805438325973228, 0.297402262222022, 0.66201988607645, 0.693481677677482, 0.415002803318202, 0.405299285659567, 0.858779930276796, 0.296279523056, 0.674106671009213, 0.578733320813626, 0.255373961059377, 0.49295330629684, 0.0747963024768978, 0.291271723341197, 0.0488648710306734, 0.511063065147027, 0.297494422411546, 0.0880304370075464, 0.0751968955155462, 0.936941209482029, 0.197233860613778, 0.985992956673726), actual = c(0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.401102338917553, 0.313667056849226, 0.0918839152436703, 0.147149522788823, 0.645829628687352, 0.127665489213541, 0.697330018505454, 0.0751103034708649, 0.576395768206567, 0.946494460105896, 0.749152758391574, 0.196900848997757, 0.650410502450541, 0.990065274061635, 0.933534340932965, 0.677132415119559, 0.373390363529325, 0.72179563739337, 0.343291723169386, 0.630000925855711, 0.994406434241682, 0.285701308399439, 0.132693767314777, 0.358284365385771, 0.514409167226404, 0.475384446559474, 0.105672085192055, 0.44457474607043, 0.842187819303945, 0.423137960489839, 0.197986118728295, 0.144100883509964, 0.504695396870375, 0.931257385294884, 0.369572212221101, 0.151859779376537, 0.170519777107984, 0.176901571918279, 0.524758796207607, 0.399150966200978, 0.683134074788541, 0.683176678139716, 0.259071211097762, 0.574881343869492, 0.0333814721088856, 0.774929133243859, 0.772918799659237, 0.603607031051069, 0.0114060617052019, 0.70439911307767, 0.927725504385307, 0.0107243014499545, 0.821676432155073, 0.41542981355451, 0.565369085641578, 0.669055945007131, 0.320536031853408, 0.0116354851052165, 0.157362104160711, 0.426098331809044, 0.136730973143131, 0.410459793871269, 0.149935188237578, 0.690222122473642, 0.910057569853961, 0.29983285209164, 0.144747441634536, 0.0997729026712477, 0.625423290301114, 0.157478392357007, 0.167653535725549, 0.00604208000004292, 0.653936031041667, 0.377846114337444, 0.383365585934371, 0.212214027764276, 0.521661009639502, 0.468417573953047, 0.348559495760128, 0.453031683573499, 0.636436682660133, 0.731890185503289, 0.349061835790053, 0.952877444913611, 0.332990732509643, 0.39549585361965, 0.269596787169576, 0.956357351969928, 0.265192380873486, 0.241725503932685, 0.257871717680246, 0.858464478747919, 0.0568773322738707, 0.602029472589493, 0.106648470973596, 0.830383093561977, 0.298351745354012, 0.186363493558019, 0.366516755893826, 0.119388099294156, 0.00716074788942933, 0.73931230022572, 0.716396698728204, 0.65676926379092, 0.792640882311389, 0.696589816128835, 0.497585156001151, 0.272643221076578, 0.375985034974292, 0.709914521779865, 0.178479193244129, 0.856353316921741, 0.31172649632208, 0.760448961751536, 0.443366107996553, 0.555080283200368, 0.557379802456126, 0.419329414144158, 0.570636434014887, 0.000807511154562235, 0.238456775434315, 0.0879481995943934, 0.694455401506275, 0.556330431019887, 0.716796792577952, 0.0232000942341983, 0.848272217903286, 0.0914430455304682, 0.144621273502707, 0.322071063797921, 0.0201825331896544, 0.00869441591203213, 0.281600117916241, 0.509248033631593, 0.000957709504291415, 0.635454541770741, 0.531267158919945, 0.293313584988937, 0.391644512303174, 0.785671758232638, 0.547229359857738, 0.514182097045705, 0.295647194841877, 0.996171162463725, 0.706839995458722, 0.0271068883594126, 0.884377185255289, 0.10502098989673, 0.178062731400132, 0.880597312701866, 0.638868056703359, 0.70791266229935, 0.327784195309505, 0.178978686453775, 0.987750796368346, 0.930979463737458, 0.359341884730384, 0.0573411383666098, 0.0834470810368657, 0.414216854143888, 0.816063646459952, 0.753095243126154, 0.440059442305937, 0.457737445132807, 0.662503038998693, 0.163144671125337, 0.968451603781432, 0.910623200936243, 0.817590553080663, 0.981139184208587, 0.526090882485732, 0.836863378761336, 0.959935751743615, 0.896624318324029, 0.390936068957672, 0.0138277483638376, 0.684059743303806, 0.753176684724167, 0.18235553544946, 0.417529092403129, 0.0892057998571545, 0.714666927233338, 0.919028029078618, 0.640772140584886, 0.800570896361023, 0.045866092434153, 0.452274084324017, 0.268183585489169, 0.763868200127035, 0.987623796565458, 0.671488802880049, 0.032832681434229, 0.925368413794786, 0.554805679246783, 0.23321899259463, 0.164596905699, 0.576333941658959, 0.825166559079662, 0.0656430465169251, 0.341849875636399, 0.410870153922588, 0.458880502497777, 0.860196534078568, 0.501426727045327, 0.511959649855271, 0.1453479337506, 0.915911299176514, 0.027745530474931, 0.607201363192871, 0.296170326415449, 0.610644784988835, 0.93389872694388, 0.365106078796089, 0.00161631661467254, 0.389746064785868, 0.241677106823772, 0.935351053019986, 0.635393274948001, 0.136929034255445, 0.463623720221221, 0.176078605698422, 0.393976389197633, 0.92663552169688, 0.394292981363833, 0.205519407987595, 0.188167904969305, 0.806173556484282, 0.660001538228244, 0.444844324840233, 0.101582149509341, 0.147749004885554, 0.704356097383425, 0.168483344139531, 0.917953344061971, 0.975696241017431, 0.807549546472728, 0.349054353777319, 0.349445292493328, 0.861814173636958, 0.149495265213773, 0.804267190629616, 0.20396028412506, 0.270118142943829, 0.963160936022177, 0.533892303705215, 0.449306530412287, 0.364178652409464, 0.673364941962063, 0.211856791749597, 0.696434626821429, 0.354210584424436, 0.360663678962737, 0.057886426569894, 0.16763771045953, 0.451634005876258, 0.996360505465418, 0.57028510957025, 0.772299079690129, 0.664195641875267, 0.851957776350901, 0.490607379702851, 0.677408007904887, 0.358265376184136, 0.0245280927047133, 0.986422176472843, 0.65534518007189, 0.362921920139343, 0.400256057502702, 0.861877407412976, 0.600777202984318, 0.5090015262831, 0.360696719260886, 0.926805175375193, 0.168150131823495, 0.564327051397413, 0.460753995692357, 0.914164198096842, 0.983972867485136, 0.375262013636529, 0.909980439813808, 0.34681821311824, 0.107771144015715, 0.755072612548247, 0.536852748598903, 0.478892360581085, 0.527594734216109, 0.215687099844217, 0.863670314894989, 0.779241020325571, 0.483750998042524, 0.669143250677735, 0.875019789440557, 0.0918982969596982, 0.912913051899523, 0.182111280271783, 0.597460061777383, 0.710626983782277, 0.251736561302096, 0.959692107979208, 0.724111540941522, 0.532022043829784, 0.976350221782923, 0.731194379972294, 0.543995629297569, 0.216272783000022, 0.596203155815601, 0.538898963714018, 0.50736815854907, 0.515115848975256, 0.176013721385971, 0.265543916495517, 0.921109998133034, 0.263728053076193, 0.899640439311042, 0.0298526489641517, 0.918513383483514, 0.651975830784068, 0.902656570076942, 0.629568746080622, 0.946479582926258, 0.581991249928251, 0.122947621857747, 0.975582563783973, 0.304475827142596, 0.235281671630219, 0.552065655589104, 0.40589367505163, 0.203164655715227, 0.757025356171653, 0.380058818496764, 0.702826650347561, 0.523692455841228, 0.626335766864941, 0.519812463317066, 0.479359333403409, 0.0546698421239853, 0.360952610848472, 0.323352206498384, 0.488348142476752, 0.0217441543936729, 0.786807262105867, 0.48023284971714, 0.3204385782592, 0.850257622776553, 0.024043204728514, 0.435947015183046, 0.135314051993191, 0.854315608041361, 0.440829069120809, 0.399394241394475, 0.490009302739054, 0.790966156404465, 0.997325743548572, 0.164240166544914, 0.702275740448385, 0.594050804618746, 0.0639093355275691, 0.88138139853254, 0.308447899995372, 0.701766269281507, 0.53501627407968, 0.946791993221268, 0.406355498358607, 0.707346031907946, 0.198828630382195, 0.344121530652046, 0.940500353230163, 0.659603605279699, 0.824186726938933, 0.661494240164757, 0.374512400710955, 0.902995759854093, 0.450736397877336, 0.151259272359312, 0.814482533140108, 0.885107792913914, 0.114175057737157, 0.36123734083958, 0.882382364245132, 0.908149518072605, 0.882112946128473, 0.636220720596611, 0.0737309521064162, 0.435820868937299, 0.0947392280213535, 0.924521065317094, 0.55237094592303, 0.18585265125148, 0.0129294339567423, 0.597424521809444, 0.1587656843476, 0.881471822271124, 0.841149497078732, 0.870220403186977, 0.827196402009577, 0.62984282290563, 0.0939551200717688, 0.634202469373122, 0.383954592049122, 0.195973211666569, 0.0572839146479964, 0.393289372324944, 0.307817655382678, 0.666778651531786, 0.273631675401703, 0.987783684162423, 0.412523043109104, 0.503033440792933, 0.287262817844748, 0.567459502257407, 0.962280530948192, 0.0546408689115196, 0.712007678579539, 0.866562578361481, 0.752829617355019, 0.388252219185233, 0.271325673209503, 0.891988915158436, 0.9618894753512, 0.952124239178374, 0.739553506951779, 0.842308320570737, 0.0842753981705755, 0.784147351514548, 0.930234586820006, 0.325204222695902, 0.909635868854821, 0.907553732162341, 0.701123547274619, 0.595097285928205, 0.304797063348815, 0.811481695389375, 0.460021421080455, 0.362696178257465, 0.701923777349293, 0.581314732087776, 0.358420245815068, 0.840093763545156, 0.412285360973328, 0.246999183902517, 0.0345850114244968, 0.554144659778103, 0.24698991002515, 0.683525819098577, 0.363570708781481, 0.651402601972222, 0.0704808961600065, 0.177134793484583, 0.237308513605967, 0.302954355953261, 0.791321133961901, 0.604724655859172, 0.261014497838914, 0.756417047930881, 0.880966467550024, 0.585476813372225, 0.0596545455045998, 0.652430944843218, 0.90705810370855, 0.954226347850636, 0.0632693816442043, 0.183131912723184, 0.330561368493363, 0.148104535415769, 0.482885911595076, 0.937621774151921, 0.432930821552873, 0.0180752011947334, 0.773131382651627, 0.768352801911533, 0.0805919296108186, 0.431764842476696, 0.23697910620831, 0.0476814296562225, 0.578305194387212, 0.683979558525607, 0.184333333512768, 0.230381779139861, 0.746314116287977, 0.107496837852523, 0.338781029451638, 0.788390960311517, 0.866534114582464, 0.919037115294486, 0.599305937532336, 0.780975003959611, 0.0106300483457744, 0.440904227085412, 0.557098529767245, 0.0796774355694652, 0.939317382406443, 0.952975737629458, 0.509845475899056, 0.754147240426391, 0.556400435743853, 0.454179015243426, 0.811252360697836, 0.723273115232587, 0.0687510611023754, 0.115033574635163, 0.450054064393044, 0.178791257319972, 0.428944858722389, 0.254856975749135, 0.425082788337022, 0.749803018989041, 0.291731007630005, 0.240215633763, 0.332354553742334, 0.610574438935146, 0.270940908929333, 0.46501295780763, 0.356600925093517, 0.989353219047189, 0.634794108802453, 0.781788499560207, 0.12343857018277, 0.038528720382601, 0.363144877599552, 0.693174394313246, 0.497957229148597, 0.76559531269595, 0.504130197223276, 0.31466947728768, 0.331371852895245, 0.438156501622871, 0.04367446876131, 0.639389982679859, 0.121031716931611, 0.162505327258259, 0.10212677763775, 0.848484349204227, 0.911725022597238, 0.86128395749256, 0.84062643838115, 0.190761429723352, 0.849965571658686, 0.909093730384484, 0.35553544270806, 0.0603844646830112, 0.0478955055586994, 0.578621159540489, 0.00511493999511003, 0.130872616777197, 0.728248343337327, 0.466301284963265, 0.26925677084364, 0.357000206597149, 0.0388865163549781, 0.0962570407427847, 0.793359491042793, 0.281000253511593, 0.0389438276179135, 0.367521868320182, 0.311456734547392, 0.202723968075588, 0.960668948013335, 0.615701924310997, 0.575623790500686, 0.0903136730194092, 0.350283834850416, 0.291691376827657, 0.129827043740079, 0.88944524503313, 0.607336859684438, 0.497601269045845, 0.143645616946742, 0.735166204627603, 0.197141252225265, 0.857410968979821, 0.166145131457597, 0.565775436814874, 0.804827492684126, 0.463977625826374, 0.705846084281802, 0.0907203054521233, 0.725500127067789, 0.839078775607049, 0.26376417488791, 0.163482914445922, 0.882467898540199, 0.141993436729535, 0.147922552889213, 0.948380863526836, 0.0809723956044763, 0.53957970184274, 0.910352294100448, 0.302600517170504, 0.214699239935726, 0.219195266021416, 0.695041853236035, 0.709869579877704, 0.585002074483782, 0.60975175560452, 0.303350295405835, 0.377745774574578, 0.489424408180639, 0.379500559298322, 0.171607010532171, 0.498644715175033, 0.695550488308072, 0.979483688250184, 0.242795260390267, 0.625756096327677, 0.472577578388155, 0.221598782110959, 0.316913542337716, 0.728984363377094, 0.198884715791792, 0.705589117016643, 0.653987656813115, 0.816975779365748, 0.0121815805323422, 0.822848056210205, 0.29297705553472, 0.343783256364986, 0.137167032342404, 0.017182843759656, 0.976422620005906, 0.891307309735566, 0.684666881104931, 0.641536032548174, 0.492924386868253, 0.260457318043336, 0.252955722622573, 0.419063478475437, 0.119236952159554, 0.444355126935989, 0.816844435175881, 0.0377676892094314, 0.34723524004221, 0.37828304246068, 0.0684813791885972, 0.719516199547797, 0.399492881260812, 0.617181564448401, 0.374866000376642, 0.222121179336682, 0.196499380050227, 0.279844788834453, 0.0566926975734532, 0.688548649428412, 0.614730548113585, 0.935242681764066, 0.337382278870791, 0.312134894076735, 0.232924499548972, 0.0540638496167958, 0.264184151077643, 0.746656160801649, 0.559994023758918, 0.174092977540568, 0.649359351489693, 0.417478417977691, 0.929347358411178, 0.698848181171343, 0.076970123918727, 0.888803673908114, 0.154863468371332, 0.581787874456495, 0.313861330971122, 0.164765830850229, 0.981937484582886, 0.196705663111061, 0.187781992601231, 0.870882706250995, 0.560776223661378, 0.248911969363689, 0.733972571557388, 0.868800114374608, 0.0892815035767853, 0.0989025642629713, 0.00404243753291667, 0.0936189719941467, 0.919682089937851, 0.62553417426534, 0.799116267822683, 0.634923377307132, 0.141010475577787, 0.716242901980877, 0.365483711939305, 0.00822027423419058, 0.374744787346572, 0.805152312852442, 0.413295489968732, 0.556336085544899, 0.212855178397149, 0.784015824086964, 0.262969841714948, 0.470021302811801, 0.141248227329925, 0.388993838569149, 0.662339741596952, 0.906215532682836, 0.265760598704219, 0.855290357489139, 0.0320554175414145, 0.597149920184165, 0.551173098850995, 0.939625212457031, 0.709173192968592, 0.0282472062390298, 0.0378816076554358, 0.249559200834483, 0.118646159768105, 0.794141603633761, 0.655275977216661, 0.101925596129149, 0.0356070762500167, 0.0427696497645229, 0.99014425300993, 0.0171742571983486, 0.966007962357253, 0.348648570710793, 0.0614567031152546, 0.567655420163646, 0.0843774192035198, 0.429637648630887, 0.228356002364308, 0.0477243852801621, 0.982963498681784, 0.764481701888144, 0.969806243665516, 0.838320634327829, 0.933121593669057, 0.727495682891458, 0.259144902927801, 0.782474601641297, 0.712967425351962, 0.741416572360322, 0.984330524457619, 0.0578503084834665, 0.695546157192439, 0.627506859833375, 0.196455074707046, 0.323941425653175, 0.74743426265195, 0.464652837952599, 0.235320776468143, 0.59152531879954, 0.0530983440112323, 0.913412864552811, 0.824184260098264, 0.956103508127853, 0.885364492656663, 0.939029677538201, 0.833722929935902, 0.0368287328165025, 0.732151719508693, 0.000418663490563631, 0.836066326824948, 0.724371046293527, 0.417430985951796, 0.493729223497212, 0.156203092541546, 0.137841294519603, 0.818421758245677, 0.508438383694738, 0.868514043046162, 0.636044425657019, 0.805301615502685, 0.429911945248023, 0.630992004880682, 0.101787704741582, 0.137067538453266, 0.502601433545351, 0.847360197687522, 0.36578783323057, 0.41459362115711, 0.642148253042251, 0.872082259273157, 0.137452027294785, 0.923135170247406, 0.969605785794556, 0.247032075654715, 0.879081927472726, 0.954907293897122, 0.190927601186559, 0.627715403446928, 0.191588607616723, 0.63832493731752, 0.864501648815349, 0.9000364090316, 0.490103112533689, 0.994727491168305, 0.436094740405679, 0.161846356233582, 0.46086327242665, 0.744781103916466, 0.415382901672274, 0.58833280345425, 0.292782322037965, 0.963680886430666, 0.19546094792895, 0.170187921496108, 0.917895136401057, 0.509173324331641, 0.476259318646044, 0.877791527658701, 0.406417882535607, 0.994282090570778, 0.998895585304126, 0.319786335807294, 0.229612255236134, 0.656572344014421, 0.428455278743058, 0.0450696907937527, 0.161638490157202, 0.441847997484729, 0.143579075112939, 0.769728721119463, 0.0325700454413891, 0.68369642063044, 0.36649108142592, 0.95031956047751, 0.116500195581466, 0.861827231943607, 0.841262989211828, 0.556472769472748, 0.816703015239909, 0.686876801773906, 0.733716450864449, 0.0346285712439567, 0.00741843436844647, 0.219169830204919, 0.632984642637894, 0.297289055772126, 0.491910116747022, 0.784919143421575, 0.0264405540656298, 0.824920323677361, 0.9854145150166, 0.55738772964105, 0.529552171006799, 0.714569153962657, 0.176912151044235, 0.918643788434565, 0.52284595509991, 0.329926442820579, 0.0604802607558668, 0.0880848951637745, 0.91371403564699, 0.523284716298804, 0.102120705880225, 0.337898274650797, 0.729818730615079, 0.339555704733357, 0.985736516071483, 0.846209198003635, 0.658235501963645, 0.827143853995949, 0.14423697325401, 0.931680996436626, 0.80657414579764, 0.592228181194514, 0.998186632059515, 0.106219191802666, 0.318254723446444, 0.655475444626063, 0.109755919314921, 0.647689457982779, 0.646507589844987, 0.543587800581008, 0.131082208594307, 0.988457377301529, 0.0841111168265343, 0.841733045177534, 0.629396175500005, 0.326802397845313, 0.954111346276477, 0.927609859965742, 0.320949483662844, 0.116592010948807, 0.336649277480319, 0.518376039573923, 0.954447818221524, 0.748791710007936, 0.937061996199191, 0.167709617875516, 0.76899242005311, 0.515136590693146, 0.314659306546673, 0.07552779186517, 0.60865147295408, 0.442539323121309, 0.0436426876112819, 0.282722887583077, 0.690232590772212, 0.872749782167375, 0.658624817617238, 0.905526392161846, 0.989004651084542, 0.143016032408923, 0.296246844343841, 0.0217644448857754, 0.67476301966235, 0.565136491553858, 0.817059176741168, 0.184638914419338, 0.889870759798214, 0.0596759049221873, 0.065846404992044, 0.393841843120754, 0.941135671688244, 0.580816281959414, 0.739998856326565, 0.493210824206471, 0.671555573586375, 0.808594916947186, 0.442690489580855, 0.690809555351734, 0.630719617241994, 0.402825516881421, 0.890561188338324, 0.595246875891462, 0.217534593539312, 0.83744063321501, 0.403214335208759, 0.414454883662984, 0.473912796238437, 0.414513190742582, 0.777255124645308, 0.94120051828213, 0.177555392030627, 0.0937963263131678, 0.725717051886022, 0.89694072981365, 0.914695980958641, 0.196579554351047, 0.774827243294567, 0.157073767390102, 0.58507394650951, 0.988511537434533, 0.806103639071807, 0.408137613441795, 0.660862229997292, 0.735102691920474, 0.327261504018679, 0.221204657573253, 0.755421164212748, 0.931029166793451, 0.103805610444397, 0.716106974286959, 0.384918726747856, 0.202417472377419, 0.996095918817446, 0.0330774560570717, 0.312027713283896, 0.903203550027683, 0.331455491483212, 0.750754706328735, 0.280253967735916, 0.805874880636111, 0.157006244407967, 0.332476705312729, 0.0462681544013321, 0.498706980841234, 0.125286442227662, 0.400313370861113, 0.56190261663869, 0.993647899245843, 0.782294597709551, 0.188049880089238, 0.325965452473611, 0.752291195094585, 0.0410674379672855, 0.392924497835338, 0.208195527084172, 0.562323616817594, 0.426315512973815, 0.209829011466354, 0.159924362553284, 0.830648390110582, 0.139211521949619, 0.0384551861789078, 0.257112054154277, 0.407767829485238, 0.140415055910125, 0.352408278035, 0.94448483036831, 0.976286338176578, 0.599164871964604, 0.502426654100418, 0.339627480134368, 0.171415970195085, 0.299956241622567, 0.699748351238668, 0.434950299328193, 0.530367644736543, 0.777944302652031, 0.0295115634799004, 0.544070773757994, 0.785449772374704, 0.707985662156716, 0.93894303473644, 0.226640830980614, 0.0543886537197977, 0.548374222125858, 0.982023612596095, 0.219894173089415), actual = c(1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"