Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.3621517        0.3807714 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4781445        0.4781445        0.5035993 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.51309279166162, 0.800320397363976, 0.679854854475707, 0.983752774074674, 0.116178234573454, 0.937172726960853, 0.9581587635912, 0.0189464020077139, 0.359211233677343, 0.85512249590829, 0.532535458914936, 0.490997652988881, 0.239084216533229, 0.911608453607187, 0.879917107289657, 0.286802675109357, 0.847150169545785, 0.903622717596591, 0.773515031440184, 0.50032346486114, 0.353670811513439, 0.66067594406195, 0.085635572206229, 0.313579735578969, 0.309373550117016, 0.262164580402896, 0.583002996863797, 0.153407433768734, 0.270813297713175, 0.355428909184411, 0.169986912515014, 0.0446800680365413, 0.295605288352817, 0.0872503009159118, 0.216153959743679, 0.290796995861456, 0.693032356211916, 0.615054313791916, 0.181079756002873, 0.382805906701833, 0.0990283880382776, 0.485276251332834, 0.505280303303152, 0.625777846667916, 0.658009513514116, 0.67047271807678, 0.976313088089228, 0.630098488880321, 0.138411399442703, 0.516822625184432, 0.440027399919927, 0.97483770805411, 0.249427521368489, 0.940407115966082, 0.467897521331906, 0.169247278477997, 0.572477129753679, 0.120631174417213, 0.712927324231714, 0.453414730029181, 0.994184178998694, 0.850660329218954, 0.150673278840259, 0.165699873352423, 0.00810893555171788, 0.9086612414103, 0.403119223192334, 0.572890280047432, 0.585990243591368, 0.362076896475628, 0.857279532589018, 0.688834559172392, 0.65177548979409, 0.804752040654421, 0.828017259947956, 0.325708098709583, 0.86002799984999, 0.762731547700241, 0.343343944987282, 0.834084202069789, 0.625441123964265, 0.842651025392115, 0.221953704953194, 0.558421763358638, 0.36240624752827, 0.165085691725835, 0.020011990563944, 0.233358065597713, 0.0466594663448632, 0.131150635657832, 0.566216390114278, 0.547018124256283, 0.939070407999679, 0.678081200225279, 0.441143498523161, 0.79487259988673, 0.289574725786224, 0.576365585438907, 0.735391115304083, 0.0729549899697304, 0.477707591839135, 0.684963846113533, 0.106977629940957, 0.152848425321281, 0.895267569227144, 0.576885813148692, 0.687332452973351, 0.809669851558283, 0.729922863189131, 0.169418375240639, 0.310092096682638, 0.0467568072490394, 0.838356537278742, 0.798395053949207, 0.681078042369336, 0.406538931885734, 0.439652831992134, 0.833817729493603, 0.70209236885421, 0.450209043687209, 0.332955332938582, 0.00869240937754512, 0.00161475175991654, 0.77415136503987, 0.962832170538604, 0.941848194226623, 0.292799829039723, 0.801038696430624, 0.828003447735682, 0.646039523649961, 0.518118400126696, 0.923318970482796, 0.951174899470061, 0.294442674145103, 0.662577972747386, 0.434286730131134, 0.519242302747443, 0.165528463898227, 0.0645979507826269, 0.633348551113158, 0.84002651181072, 0.657281886087731, 0.172362134791911, 0.870436755241826, 0.326223129639402, 0.887235966976732, 0.761457574553788, 0.0220986914355308, 0.832307477481663, 0.421247072517872, 0.833921636454761, 0.382191999815404, 0.176167542580515, 0.778888884931803, 0.549028465291485, 0.633869953686371, 0.986202332656831, 0.778621977893636, 0.187344741309062, 0.56821068469435, 0.849043956492096, 0.656348030781373, 0.470207214588299, 0.159597814781591, 0.0197633863426745, 0.520330791128799, 0.381354548735544, 0.851773741422221, 0.27350892056711, 0.54085446940735, 0.140324430307373, 0.0887636684346944, 0.554974881000817, 0.306726428447291, 0.669629582669586, 0.905925090657547, 0.130842542508617, 0.438213280867785, 0.198269868036732, 0.474905815906823, 0.875023754313588, 0.609465608373284, 0.195417237235233, 0.569530798355117, 0.798170084599406, 0.59830296295695, 0.412800332298502, 0.545608642278239, 0.648149655666202, 0.0137345928233117, 0.652963472064584, 0.556237522512674, 0.331811643671244, 0.458225300069898, 0.146970909787342, 0.868550850544125, 0.0664898985996842, 0.642768056830391, 0.292588036041707, 0.123533834703267, 0.811076180543751, 0.937173013575375, 0.296166100306436, 0.395864058518782, 0.470857643987983, 0.474746498512104, 0.541657746303827, 0.44432895956561, 0.47566531971097, 0.790517932502553, 0.590329471975565, 0.894248016644269, 0.107919494621456, 0.530172877712175, 0.218078669160604, 0.89038325753063, 0.380417519016191, 0.591238493099809, 0.617668407503515, 0.120695468271151, 0.737758234608918, 0.613543856889009, 0.647626594640315, 0.124888117890805, 0.253471335861832, 0.836430521449074, 0.828120407415554, 0.735738314455375, 0.832706319866702, 0.974787995452061, 0.437436850508675, 0.42884506797418, 0.281121166888624, 0.677968845935538, 0.821861732052639, 0.884173163212836, 0.996550909243524, 0.0741492023225874, 0.837302017956972, 0.192604863317683, 0.982476201141253, 0.541196447331458, 0.87517975573428, 0.667097628582269, 0.929177542449906, 0.652108424343169, 0.349522332660854, 0.094199943356216, 0.8445093925111, 0.436961779836565, 0.0923056751489639, 0.201568501535803, 0.240058400202543, 0.968708941945806, 0.855259774019942, 0.58315549348481, 0.581787907984108, 0.526275170966983, 0.918777559883893, 0.477071906905621, 0.944347252137959, 0.480564034311101, 0.914289466571063, 0.504473927663639, 0.946348795667291, 0.650959020946175, 0.743136889534071, 0.501119916094467, 0.730100702028722, 0.417916740290821, 0.0596579865086824, 0.533222459256649, 0.918600933160633, 0.990757199935615, 0.0312203101348132, 0.215484753716737, 0.987153316149488, 0.908292644191533, 0.844205210218206, 0.258952979231253, 0.514972054865211, 0.904640240129083, 0.156304432544857, 0.14354656287469, 0.0461771241389215, 0.592034864239395, 0.63329590880312, 0.217185123823583, 0.355640288675204, 0.897650041384622, 0.916189764859155, 0.324304334586486, 0.413708804640919, 0.969356305664405, 0.429802522994578, 0.29099377989769, 0.382963345153257, 0.77744260779582, 0.919355754740536, 0.680556301726028, 0.0276847262866795, 0.275835000444204, 0.44060286344029, 0.310899797827005, 0.355458174599335, 0.164221700048074, 0.0320150637999177, 0.858279404463246, 0.375059478683397, 0.131382205989212, 0.759850368369371, 0.584698268678039, 0.257610198343173, 0.265559777151793, 0.747731405543163, 0.739079259103164, 0.622346464078873, 0.350689421407878, 0.185014493064955, 0.0606310442090034, 0.5202001912985, 0.369832293828949, 0.664948647608981, 0.895545041654259, 0.567254953086376, 0.782410711515695, 0.288651281967759, 0.480576424859464, 0.695104593411088, 0.589793758932501, 0.0641622522380203, 0.391197813209146, 0.260935353813693, 0.820679250871763, 0.243747019208968, 0.443767678225413, 0.576561777153984, 0.183730828110129, 0.422692499123514, 0.969444890273735, 0.780441951239482, 0.0755418208427727, 0.833537959260866, 0.539597045863047, 0.0872253391426057, 0.553100514691323, 0.0760928308591247, 0.0477736280299723, 0.857229408808053, 0.89827647455968, 0.763512025587261, 0.658543424447998, 0.906601553782821, 0.532431277213618, 0.448468494461849, 0.859297487651929, 0.568742267321795, 0.138824470806867, 0.239899902138859, 0.426389102591202, 0.485783596988767, 0.163927091052756, 0.964640631806105, 0.171848315978423, 0.0609986176714301, 0.377845128066838, 0.608696223469451, 0.940289611928165, 0.667105600936338, 0.503226949600503, 0.574712674599141, 0.886942876502872, 0.0262723306659609, 0.848193803569302, 0.493491172092035, 0.0300479582510889, 0.103704075794667, 0.572161940159276, 0.544118742225692, 0.48029196029529, 0.202213302953169, 0.948867316124961, 0.499409317504615, 0.617819218430668, 0.142615551594645, 0.625891375355422, 0.307293390389532, 0.127358985133469, 0.690300924237818, 0.377862963825464, 0.670503240078688, 0.480628723977134, 0.190320161869749, 0.126777133671567, 0.256474579451606, 0.604321343591437, 0.271771864034235, 0.688559180824086, 0.174594376469031, 0.639638579217717, 0.350069903535768, 0.368732224684209, 0.736275842646137, 0.428695210488513, 0.816951486980543, 0.53709813579917, 0.14121055486612, 0.102907544234768, 0.334383979905397, 0.00807152828201652, 0.654628819320351, 0.173760333564132, 0.656932427780703, 0.816808190429583, 0.663150410400704, 0.893294388893992, 0.397057798691094, 0.488197623053566, 0.853122272528708, 0.769612910225987, 0.320230329642072, 0.326698941411451, 0.447426954749972, 0.205999451689422, 0.656306434189901, 0.8807592112571, 0.210468239616603, 0.675944867078215, 0.45598234445788, 0.327510010218248, 0.810254641342908, 0.754265064373612, 0.868902221089229, 0.313367954688147, 0.455843278439716, 0.156119666760787, 0.531308945268393, 0.254857176914811, 0.692571373656392, 0.423628683201969, 0.78769665164873, 0.0779356730636209, 0.391601908486336, 0.797141022514552, 0.406300623202696, 0.605557672912255, 0.0141133610159159, 0.565838532755151, 0.381302035646513, 0.131012756610289, 0.520381138660014, 0.343656469834968, 0.374273632187396, 0.465318626491353, 0.0968062523752451, 0.501188582740724, 0.648347774054855, 0.464114502537996, 0.408320672344416, 0.260263256495818, 0.106371814850718, 0.654271422186866, 0.0576410330832005, 0.670084882294759, 0.423667937749997, 0.721047561382875, 0.0210524511057884, 0.0390608829911798, 0.552760540274903, 0.674351308494806, 0.182868709322065, 0.831854362972081, 0.491416421951726, 0.775527608580887, 0.932469984283671, 0.121269485214725, 0.0385449503082782, 0.813247109996155, 0.857571620494127, 0.91210352955386, 0.30391092877835, 0.551787720294669, 0.558214432559907, 0.253955492051318, 0.689564066007733, 0.567621573572978, 0.516878305468708, 0.620193949900568, 0.869918416487053, 0.456841748440638, 0.141094092279673, 0.50845142826438, 0.761113729095086, 0.68101829662919, 0.836488932138309, 0.303581986343488, 0.776322984136641, 0.846694026142359, 0.104869297239929, 0.190636159153655, 0.591843907022849, 0.692704929737374, 0.0575737424660474, 0.757757489569485, 0.5024884424638, 0.723414656706154, 0.00138146453537047, 0.234729560790583, 0.148801181465387, 0.521634328877553, 0.352097062394023, 0.449974513379857, 0.798596766078845, 0.969676712760702, 0.778100844239816, 0.521839095512405, 0.31522338045761, 0.620804356876761, 0.00193346990272403, 0.768524384824559, 0.563206410501152, 0.0192623983602971, 0.424516304396093, 0.883991015376523, 0.0135919502936304, 0.192934083752334, 0.72199254273437, 0.449492195155472, 0.331439204048365, 0.00437142024748027, 0.774243853054941, 0.025250005768612, 0.424471772043034, 0.380871093366295, 0.6933353994973, 0.844953762367368, 0.804826990002766, 0.835804372560233, 0.727367578307167, 0.55739917862229, 0.314531408948824, 0.0398003428708762, 0.153571991948411, 0.944497426040471, 0.860571276629344, 0.340948721626773, 0.112076516263187, 0.567339757224545, 0.445787785574794, 0.261798833496869, 0.311066889436916, 0.373083976563066, 0.19614613451995, 0.772433900507167, 0.04236776009202, 0.674307553330436, 0.307899824110791, 0.543659221148118, 0.0957686151377857, 0.846845245687291, 0.830991399474442, 0.657778725959361, 0.0602648619096726, 0.412880070507526, 0.479899150319397, 0.78385609574616, 0.362594619393349, 0.00906576937995851, 0.396147882798687, 0.198548441752791, 0.451662274077535, 0.871389929205179, 0.754066092893481, 0.354666517581791, 0.7629250632599, 0.527148026274517, 0.606973313260823, 0.517600395483896, 0.458735679974779, 0.613669269019738, 0.125722226919606, 0.409561906242743, 0.962089933920652, 0.743266406701878, 0.988709490280598, 0.37462476012297, 0.703567264601588, 0.079369448358193, 0.212334920186549, 0.0295696696266532, 0.0682937840465456, 0.827334117842838, 0.676699811825529, 0.864127584500238, 0.869597868993878, 0.0312604343052953, 0.982046940363944, 0.673146776622161, 0.0475820922292769, 0.305708861909807, 0.797353103291243, 0.145565397804603, 0.0536982680205256, 0.407807066338137, 0.306014889385551, 0.033362629590556, 0.544880104949698, 0.114748310763389, 0.165072632255033, 0.483443292789161, 0.130889989435673, 0.853524989914149, 0.791830468457192, 0.192162432707846, 0.0884965984150767, 0.621832413366064, 0.527024475159124, 0.529392403550446, 0.412435393314809, 0.36012823251076, 0.460834451019764, 0.922336899442598, 0.830755835399032, 0.0741694699972868, 0.00605833018198609, 0.83446881477721, 0.557769751176238, 0.50938193523325, 0.620156277203932, 0.258939799154177, 0.932084576692432, 0.43348868121393, 0.148831165628508, 0.271570244338363, 0.997850578743964, 0.417214586632326, 0.820482796989381, 0.136436582310125, 0.49579956009984, 0.356101311277598, 0.629114828072488, 0.920149512123317, 0.278559726197273, 0.641583818709478, 0.450884920312092, 0.427756167715415, 0.190051381010562, 0.166009625187144, 0.594320942182094, 0.30243788077496, 0.693138332106173, 0.176625607768074, 0.346685810480267, 0.698482438689098, 0.46583681832999, 0.149632581043988, 0.386188595555723, 0.370760515099391, 0.157842827960849, 0.64470605738461, 0.977844575420022, 0.789733670419082, 0.5943631043192, 0.673702738014981, 0.0819969105068594, 0.514674256555736, 0.42429359909147, 0.350470960140228, 0.395092878490686, 0.716638875659555, 0.282047452172264, 0.832464501727372, 0.372427574591711, 0.599102443782613, 0.41259344201535, 0.368467699503526, 0.219271318055689, 0.885219198418781, 0.369445827091113, 0.664526254404336, 0.470349667593837, 0.432764241937548, 0.116468837251887, 0.120146895293146, 0.075614859117195, 0.206473454833031, 0.579424598952755, 0.0238512919750065, 0.134919845033437, 0.612727184779942, 0.108285168651491, 0.608967592706904, 0.508815028471872, 0.132710409816355, 0.920906859217212, 0.899209803435951, 0.381683959858492, 0.325685647083446, 0.421099256491289, 0.396802644478157, 0.457784253405407, 0.724704529391602, 0.959250967716798, 0.914240695536137, 0.636452676495537, 0.378393068676814, 0.611000161617994, 0.383598199347034, 0.160562408389524, 0.0412107470910996, 0.746128681348637, 0.324292342877015, 0.890426333528012, 0.989900124259293, 0.796352358069271, 0.480074790306389, 0.927586006466299, 0.372930794022977, 0.59698143764399, 0.724170628935099, 0.990232338197529, 0.489404748659581, 0.936451928922907, 0.8336476017721, 0.442875273525715, 0.636768009047955, 0.235934692667797, 0.239626552909613, 0.226178905228153, 0.635474364506081, 0.00161470053717494, 0.413456537527964, 0.71606694557704, 0.508221199270338, 0.764760883292183, 0.655865618027747, 0.317739834776148, 0.989003035705537, 0.0547224506735802, 0.140827192924917, 0.0951782923657447, 0.21790999523364, 0.634342805016786, 0.362196972360834, 0.0581627918872982, 0.612944236490875, 0.154370936565101, 0.370661901542917, 0.295422993134707, 0.0827005484607071, 0.681344049051404, 0.251009551342577, 0.828508253674954, 0.970775438006967, 0.557503573596478, 0.3662409032695, 0.356137918075547, 0.104221173096448, 0.638045134488493, 0.140719221439213, 0.341692239977419, 0.367455716012046, 0.69386302260682, 0.809627438895404, 0.437474603531882, 0.562063972698525, 0.987609480274841, 0.538189945509657, 0.866820117691532, 0.407742528477684, 0.353146048262715, 0.405621414538473, 0.154809445142746, 0.532823326298967, 0.595288832904771, 0.776257278164849, 0.297245040070266, 0.542566610034555, 0.201365213142708, 0.331348426872864, 0.309831988066435, 0.971537210978568, 0.783816138282418, 0.570292522665113, 0.637948374729604, 0.96210834546946, 0.85046144365333, 0.877320799976587, 0.565114488825202, 0.880127875134349, 0.0457504924852401, 0.635981525992975, 0.419716482982039, 0.980474854353815, 0.825722432229668, 0.1189497965388, 0.923389340052381, 0.783519518096, 0.138772977283224, 0.601100781932473, 0.20045772427693, 0.279065381735563, 0.260358858853579, 0.542567256372422, 0.750796441687271, 0.957996521843597, 0.0511586994398385, 0.0755951160099357, 0.516905820462853, 0.322266313713044, 0.578975835815072, 0.284392670728266, 0.0286142986733466, 0.546909796074033, 0.906200216850266, 0.382549026748165, 0.537318034330383, 0.13762282836251, 0.602851050440222, 0.850432938197628, 0.427156443241984, 0.0689503559842706, 0.84816364152357, 0.570198661880568, 0.522989901946858, 0.539769778028131, 0.286135650705546, 0.544168442720547, 0.862771082436666, 0.0381706324405968, 0.270291054621339, 0.611787305213511, 0.988353739958256, 0.277948232833296, 0.955307650379837, 0.836647347779945, 0.30501128709875, 0.456735994666815, 0.380013200454414, 0.763537670020014, 0.015121208736673, 0.934219854185358, 0.917560677742586, 0.77223145053722, 0.639224216109142, 0.785565636819229, 0.167531138518825, 0.446576236514375, 0.614338275510818, 0.0836990566458553, 0.0604236477520317, 0.488787800306454, 0.773554068757221, 0.0726360939443111, 0.922427507583052, 0.223903717240319, 0.132701845606789, 0.281576635083184, 0.10793535085395, 0.784930641297251, 0.284477722365409, 0.722729566507041, 0.980887359473854, 0.825419769855216, 0.634793212870136, 0.330776791553944, 0.769678787793964, 0.93360375193879, 0.409553700825199, 0.658249642932788, 0.0932237121742219, 0.0508086422923952, 0.753869423409924, 0.372404588852078, 0.508509895764291, 0.0698304837569594, 0.132675553206354, 0.00130032771266997, 0.679148464463651, 0.269506519660354, 0.573551174718887, 0.905484744114801, 0.868010271107778, 0.969303969293833, 0.147174075944349, 0.841271300334483, 0.666273835115135, 0.0818172313738614, 0.79517849558033, 0.806099039036781, 0.0404531396925449, 0.497330129379407, 0.855250673601404, 0.562914722366259, 0.804213596042246, 0.168190186144784, 0.0471122623421252, 0.368229245068505, 0.235278684413061, 0.980847426922992, 0.282414465676993, 0.991701380349696, 0.956002895953134, 0.63359209895134, 0.69397765933536, 0.0737957491073757, 0.217389651341364, 0.678807134507224, 0.0585487149655819, 0.912505890941247, 0.928904610918835, 0.316415256820619, 0.259962317533791, 0.684726026374847, 0.462422676151618, 0.859069210244343, 0.902914650738239, 0.365565242711455, 0.0396621767431498, 0.150014387909323, 0.254744912264869, 0.973510883050039, 0.428440226707608, 0.621068990789354, 0.0804815827868879, 0.49372828938067, 0.570226821117103, 0.874478749930859, 0.0140045187436044, 0.705807059304789, 0.484809614950791, 0.0376401944085956, 0.347351332195103, 0.848910722648725, 0.981675915420055, 0.646006266819313, 0.0970960664562881, 0.731581722386181, 0.213909080019221, 0.836823059245944, 0.651268118759617, 0.687673216219991, 0.266184949316084, 0.762694970006123, 0.635137005476281, 0.0546787998173386, 0.788080337224528, 0.0739884022623301, 0.557924093678594, 0.0601789630018175, 0.477649880340323, 0.332947968738154, 0.376744838431478, 0.499221688136458, 0.0766649749130011, 0.787698772968724, 0.28551885811612, 0.972498697927222, 0.558504730463028, 0.219792265910655, 0.033913761144504, 0.999689214862883, 0.116518945433199, 0.21607887535356, 0.905678713927045, 0.1674003158696, 0.512285568518564, 0.137189176399261, 0.665140580385923, 0.959940015338361, 0.906508641084656, 0.360681670019403, 0.293094196356833, 0.983123777201399, 0.474499709438533, 0.553368500201032, 0.745065873255953, 0.0976069166790694, 0.228621510323137, 0.107270158827305, 0.949028718052432, 0.138792231446132, 0.680476018693298, 0.443608447909355, 0.607781933853403, 0.778961601899937, 0.353727994952351, 0.702174538513646, 0.633012032136321, 0.274706140859053, 0.426229779142886, 0.304049883270636, 0.327610175125301, 0.353624462150037, 0.36365367192775, 0.925116138998419, 0.377145334146917, 0.124063706956804, 0.962772995233536, 0.608962436905131), actual = c(0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.952067801263183, 0.162234830437228, 0.74006800330244, 0.654037727043033, 0.183800536207855, 0.69910957175307, 0.298993125557899, 0.120130160823464, 0.966293003410101, 0.0204256426077336, 0.696053233463317, 0.0209663545247167, 0.245471058180556, 0.97929470310919, 0.848237518453971, 0.791167142335325, 0.399557684082538, 0.093610221054405, 0.293238994665444, 0.618205340346321, 0.323036811081693, 0.142832509241998, 0.187838565558195, 0.828365030232817, 0.887067078147084, 0.862928384216502, 0.15105296485126, 0.933370679151267, 0.812314889393747, 0.789876708062366, 0.902204972459003, 0.338153145508841, 0.500842737033963, 0.757094501052052, 0.622310901992023, 0.404042290290818, 0.544334402075037, 0.202163067646325, 0.458884582854807, 0.273877217667177, 0.855875123292208, 0.78746857191436, 0.760537731926888, 0.570067455526441, 0.566390730906278, 0.00872019748203456, 0.881990093272179, 0.612217579502612, 0.53557861642912, 0.552811225410551, 0.0353355424012989, 0.64043631660752, 0.322961534140632, 0.89535925607197, 0.833748708944768, 0.677161188796163, 0.692915411200374, 0.683566903695464, 0.641620664624497, 0.808940967312083, 0.723295927513391, 0.255862410413101, 0.86223075655289, 0.812387467827648, 0.101428114809096, 0.44509448716417, 0.898998703574762, 0.981264765374362, 0.518962704809383, 0.811604848364368, 0.764818220166489, 0.922493340447545, 0.574017821112648, 0.619274136377499, 0.373953601112589, 0.294943558517843, 0.246495918603614, 0.965365161420777, 0.358725167345256, 0.678319116355851, 0.442256874404848, 0.87077491171658, 0.581511288415641, 0.563937954138964, 0.384783268673345, 0.987319387029856, 0.521840278757736, 0.835510378936306, 0.932924738619477, 0.124761956045404, 0.158134669298306, 0.906598720932379, 0.97000716230832, 0.0528874020092189, 0.203076386358589, 0.187813720433041, 0.320876196492463, 0.328102430561557, 0.994001179933548, 0.264357163570821, 0.758418135577813, 0.0152737742755562, 0.74081166787073, 0.878888773731887, 0.107730624964461, 0.612883323803544, 0.728559751063585, 0.250839614542201, 0.666178750572726, 0.406116123311222, 0.643221275415272, 0.31643927632831, 0.923734205309302, 0.973360309842974, 0.375296499114484, 0.518647773656994, 0.232287108199671, 0.0168415072839707, 0.795277514262125, 0.995586932171136, 0.717602469492704, 0.0507265161722898, 0.0181617781054229, 0.219715398969129, 0.795133635867387, 0.189031479647383, 0.455586850643158, 0.392324562184513, 0.553500424139202, 0.15620961971581, 0.0791302977595478, 0.708703204290941, 0.398878040025011, 0.496503752656281, 0.566582914208993, 0.67596549144946, 0.656798353185877, 0.0912369871512055, 0.0341587816365063, 0.951244916534051, 0.336300816386938, 0.934923083288595, 0.266405123285949, 0.916574166156352, 0.0054857162758708, 0.427197243086994, 0.311748327920213, 0.608434800058603, 0.45843344531022, 0.615131559083238, 0.722660110564902, 0.0345266540534794, 0.657879224047065, 0.385971608106047, 0.97566089220345, 0.290522383758798, 0.639205895829946, 0.610752435866743, 0.0689461536239833, 0.736795787233859, 0.399183472851291, 0.0317502147518098, 0.287156510865316, 0.0361347086727619, 0.134848536457866, 0.0362305340822786, 0.945664173457772, 0.793157025240362, 0.694851265754551, 0.734527589287609, 0.338358722394332, 0.620581143302843, 0.774034428410232, 0.732232889626175, 0.973132202168927, 0.326545200077817, 0.628030158346519, 0.113089259946719, 0.602835454978049, 0.183818462304771, 0.326270144898444, 0.158884920645505, 0.281064653536305, 0.488115963991731, 0.86109436978586, 0.609593977918848, 0.91428160155192, 0.146718555130064, 0.71421998552978, 0.272073997650295, 0.0799123579636216, 0.822007630486041, 0.513567339396104, 0.162617680849507, 0.0400262740440667, 0.38325161463581, 0.0765673210844398, 0.947424048092216, 0.56790113216266, 0.285987702663988, 0.702666676836088, 0.817639978136867, 0.899441299960017, 0.480452150339261, 0.650622174143791, 0.398754682624713, 0.115914584137499, 0.04092165408656, 0.573327966965735, 0.171635187463835, 0.80015442869626, 0.36117173708044, 0.755650171544403, 0.166034457273781, 0.0244484385475516, 0.46709172334522, 0.647545140935108, 0.412960881832987, 0.525913928868249, 0.106604917207733, 0.295452023623511, 0.607573206536472, 0.628195010591298, 0.195432977518067, 0.57162676518783, 0.635294170351699, 0.348360150586814, 0.74639784055762, 0.397711512632668, 0.481132280547172, 0.00916714430786669, 0.682040294399485, 0.168477069819346, 0.374823409831151, 0.297522262204438, 0.104911508969963, 0.170412180945277, 0.44990378851071, 0.0471954932436347, 0.680179866030812, 0.839350970461965, 0.636946277925745, 0.416950500570238, 0.484081639675424, 0.958300721365958, 0.396674498450011, 0.600572443800047, 0.440776840783656, 0.849001849303022, 0.631921384949237, 0.0996329227928072, 0.776847205357626, 0.0390248312614858, 0.269026687601581, 0.929833995644003, 0.777309960220009, 0.947731415275484, 0.774811159353703, 0.0333539007697254, 0.25297755212523, 0.207720337901264, 0.630329186562449, 0.636569355148822, 0.232127142837271, 0.211380009073764, 0.887750297784805, 0.420871226349846, 0.395183856599033, 0.522505866596475, 0.176668782252818, 0.530423794873059, 0.201249184086919, 0.374474978074431, 0.625209306133911, 0.590783288935199, 0.468570863828063, 0.459943236550316, 0.810098386369646, 0.998997495044023, 0.595211821142584, 0.345520244445652, 0.209493691334501, 0.941349151777104, 0.940808375598863, 0.153791507706046, 0.760164486244321, 0.839787072036415, 0.709526627790183, 0.746182941133156, 0.757410020334646, 0.868293298408389, 0.578326312592253, 0.620214022696018, 0.255675318418071, 0.763885657303035, 0.360319372965023, 0.701636693207547, 0.14710838184692, 0.91881363443099, 0.596035480732098, 0.745778556913137, 0.0541665595956147, 0.118512518471107, 0.189144703093916, 0.7852780201938, 0.731470780912787, 0.782012298470363, 0.952244880143553, 0.292270179837942, 0.639413029421121, 0.0255055138841271, 0.398659603670239, 0.0252783510368317, 0.109541211975738, 0.845989820314571, 0.342232947703451, 0.423493782989681, 0.785592456581071, 0.774687965633348, 0.921932724770159, 0.156864648219198, 0.338498251745477, 0.159594478318468, 0.550362545764074, 0.0824230723083019, 0.370020655216649, 0.643697079271078, 0.759836288401857, 0.248989259824157, 0.441037338925526, 0.583792156074196, 0.258333666017279, 0.997869758866727, 0.399966274388134, 0.0429694002959877, 0.00490115093998611, 0.694085519062355, 0.116354197729379, 0.671703767031431, 0.334928843425587, 0.875961155630648, 0.0791845147032291, 0.216822128975764, 0.32376798777841, 0.138954033376649, 0.13048692420125, 0.0799160930328071, 0.853952414123341, 0.0124700965825468, 0.605881557799876, 0.872308233287185, 0.240902819205076, 0.896592948352918, 0.0653265593573451, 0.949397419579327, 0.443247765535489, 0.0865785928908736, 0.98989431024529, 0.65784098347649, 0.129264393122867, 0.0493939795996994, 0.614144253078848, 0.362425832776353, 0.980901312548667, 0.878503866959363, 0.804598192451522, 0.504840499488637, 0.970864292234182, 0.284123696852475, 0.500370448688045, 0.447564062895253, 0.826151778455824, 0.714665325125679, 0.216648461297154, 0.911906572757289, 0.652087640017271, 0.308948983671144, 0.474619484739378, 0.440938753075898, 0.596538004931062, 0.817374372156337, 0.444415270583704, 0.0982879321090877, 0.702552784234285, 0.104532725876197, 0.963857346447185, 0.945434619672596, 0.588614110369235, 0.126119782216847, 0.0467556680087, 0.956527469446883, 0.56174828694202, 0.841802916256711, 0.0130080580711365, 0.618460290133953, 0.897504977649078, 0.887580097652972, 0.424302560975775, 0.750113004352897, 0.56562076206319, 0.959139668382704, 0.352179160807282, 0.95357171096839, 0.183795972028747, 0.600266739260405, 0.886130610248074, 0.672524909954518, 0.668825826374814, 0.746505057439208, 0.847425767453387, 0.594807643909007, 0.368832466658205, 0.565145135158673, 0.593839165987447, 0.501208285102621, 0.794155138777569, 0.668470232281834, 0.354350084904581, 0.125728321494535, 0.328220335068181, 0.940852841362357, 0.198153848294169, 0.325091961538419, 0.715374916559085, 0.682945092208683, 0.64856289816089, 0.413612683769315, 0.905273783719167, 0.170864329906181, 0.627331434283406, 0.289648922625929, 0.491445200983435, 0.279351066565141, 0.348370330641046, 0.842180358013138, 0.682249165372923, 0.961019632406533, 0.0868407024536282, 0.17198902531527, 0.0172601079102606, 0.96783263515681, 0.53390759951435, 0.262868351303041, 0.988331381930038, 0.297254721866921, 0.967954424908385, 0.9775803796947, 0.973916590213776, 0.987080293474719, 0.969528324203566, 0.420951527077705, 0.199422062141821, 0.952651948668063, 0.0164067069999874, 0.399637879105285, 0.371056236559525, 0.691992371808738, 0.442539874464273, 0.779225274454802, 0.536805785493925, 0.638100752839819, 0.425752073759213, 0.00158828194253147, 0.080215192399919, 0.301048394059762, 0.22664190758951, 0.248160852584988, 0.761353878537193, 0.791202431777492, 0.439596121199429, 0.319787759333849, 0.469165557762608, 0.138841263018548, 0.642296459991485, 0.837444209959358, 0.718474137131125, 0.529509399784729, 0.542799077229574, 0.548856984591112, 0.401350517524406, 0.725328281754628, 0.522639255272225, 0.464937052223831, 0.585015377262607, 0.0304797329008579, 0.327003636630252, 0.620474275201559, 0.781722057145089, 0.443302302854136, 0.896436082199216, 0.29422203428112, 0.322357740951702, 0.815251361578703, 0.328985251486301, 0.312135319691151, 0.574273398844525, 0.422145897289738, 0.138727396028116, 0.946218218654394, 0.202322933357209, 0.796884800074622, 0.396012340905145, 0.182558171916753, 0.130594917107373, 0.183000867487863, 0.570350166643038, 0.235534150153399, 0.971585084218532, 0.261099988827482, 0.609990780940279, 0.0699688827153295, 0.0631170833949, 0.685653618304059, 0.767216704087332, 0.912071242695674, 0.239888950018212, 0.489453879883513, 0.985819746507332, 0.239533090265468, 0.489565424621105, 0.696322183590382, 0.482648163102567, 0.520763816079125, 0.522773296805099, 0.252965435618535, 0.392604812048376, 0.51471936982125, 0.398413961054757, 0.0706144582945853, 0.378658103058115, 0.196831886423752, 0.2919734318275, 0.144085761392489, 0.488959941314533, 0.34496647096239, 0.245869291480631, 0.00928304647095501, 0.170679928502068, 0.165318392915651, 0.386806041933596, 0.919607733609155, 0.683555989526212, 0.128455615136772, 0.532514329534024, 0.86166546237655, 0.669733411399648, 0.11369768762961, 0.199569855351001, 0.698126627365127, 0.395764312241226, 0.151014923350886, 0.563073036260903, 0.767096434952691, 0.150735501199961, 0.586830246029422, 0.684210160980001, 0.114307940937579, 0.867165501927957, 0.0354327254462987, 0.16420056251809, 0.701184252044186, 0.748714473797008, 0.576915802201256, 0.324758142465726, 0.647464691195637, 0.873578213853762, 0.473614630987868, 0.614120773971081, 0.912471604300663, 0.138013647403568, 0.485444832127541, 0.579817857593298, 0.355639898916706, 0.446126231923699, 0.356167485006154, 0.196782481390983, 0.945241190260276, 0.730261448537931, 0.561157587449998, 0.618482551304623, 0.735131492372602, 0.237076268997043, 0.476998568279669, 0.376808508299291, 0.264181103790179, 0.280730181606486, 0.272133720805869, 0.962698340183124, 0.37240797909908, 0.288068321999162, 0.12020226335153, 0.488607517443597, 0.791432598372921, 0.520850271685049, 0.880147563759238, 0.272961723152548, 0.0102252296637744, 0.452455047750846, 0.783787782303989, 0.606508932774886, 0.499515023548156, 0.955754760187119, 0.926990726729855, 0.853549013379961, 0.0540615026839077, 0.843233957188204, 0.599328498356044, 0.667108939262107, 0.581715862732381, 0.365821689832956, 0.0343784696888179, 0.45344131370075, 0.546968988608569, 0.0863241909537464, 0.843999188626185, 0.861570016713813, 0.86314877262339, 0.816505960887298, 0.129434758564457, 0.144376385724172, 0.327553644077852, 0.279598127352074, 0.397174222394824, 0.642568063223734, 0.404098769649863, 0.233485841890797, 0.404563327552751, 0.188688192982227, 0.163555582053959, 0.432256284402683, 0.141893946100026, 0.161175243090838, 0.156114741694182, 0.745977888815105, 0.953509256010875, 0.338668931974098, 0.605217046104372, 0.288325448753312, 0.238969054771587, 0.566623158752918, 0.949626952176914, 0.538752695545554, 0.18972509750165, 0.402853580191731, 0.888507466530427, 0.948055075481534, 0.241024157963693, 0.471550053916872, 0.783590960549191, 0.836290017701685, 0.631462243851274, 0.797747520264238, 0.781993094598874, 0.114791330182925, 0.240921014919877, 0.810832240851596, 0.442169179907069, 0.0068829741794616, 0.687249141745269, 0.850382874254137, 0.755281619494781, 0.296329696197063, 0.958899437915534, 0.989828893681988, 0.759887741645798, 0.563164678402245, 0.127229209523648, 0.0317986221052706, 0.824961345177144, 0.823070180602372, 0.116961177671328, 0.232940031215549, 0.489597680745646, 0.105247987201437, 0.435127021279186, 0.687346010003239, 0.794113767333329, 0.211264431010932, 0.969931974075735, 0.168106829747558, 0.252274104394019, 0.924614937510341, 0.555377369048074, 0.962929350091144, 0.799491192679852, 0.471785177476704, 0.263734689680859, 0.575033306609839, 0.544646499212831, 0.700095941312611, 0.588335752487183, 0.664506408385932, 0.677375777391717, 0.630203411681578, 0.652492778608575, 0.00531420460902154, 0.706956289708614, 0.656761585269123, 0.371430603787303, 0.738129884470254, 0.394382555736229, 0.126305689802393, 0.141577476868406, 0.63375212950632, 0.337022559717298, 0.341103760059923, 0.903134589083493, 0.314692949876189, 0.716666883789003, 0.133066930808127, 0.940034805098549, 0.0914823727216572, 0.620668001938611, 0.930267081363127, 0.285389872500673, 0.248706092359498, 0.375808448530734, 0.00674975872971117, 0.644060180755332, 0.851538674673066, 0.867651192937046, 0.618991436203942, 0.430135249625891, 0.0299378775525838, 0.863794676028192, 0.213115365710109, 0.167965960456058, 0.340134497499093, 0.957241454627365, 0.228324750671163, 0.0545227043330669, 0.45773736294359, 0.728115655481815, 0.218788012163714, 0.319943215232342, 0.415054868673906, 0.180378132965416, 0.438243913464248, 0.865865652449429, 0.590496042976156, 0.708166426047683, 0.394330603536218, 0.644605275942013, 0.032841894775629, 0.376399833941832, 0.561970404116437, 0.901085362536833, 0.977465490112081, 0.858729190425947, 0.326378519879654, 0.902660462539643, 0.158494380535558, 0.102454448351637, 0.378113570157439, 0.888857346726581, 0.100637405877933, 0.381202738266438, 0.771228568162769, 0.15496864868328, 0.621800671564415, 0.0989929833449423, 0.508643380831927, 0.0778436360415071, 0.364798374008387, 0.583138695685193, 0.682795041473582, 0.683490262366831, 0.57967774476856, 0.46972206630744, 0.477613420924172, 0.355773288290948, 0.354088456602767, 0.631014067213982, 0.961785505060107, 0.457473520655185, 0.171151715097949, 0.5392592754215, 0.370133702177554, 0.636433357605711, 0.534769279882312, 0.191750272642821, 0.89291816111654, 0.577518660342321, 0.585385695332661, 0.699434788199142, 0.977335520554334, 0.464914352633059, 0.053636206779629, 0.66904324805364, 0.21237885649316, 0.103386961389333, 0.983892762102187, 0.580982412677258, 0.265253081219271, 0.64591852389276, 0.914540958357975, 0.563419840997085, 0.839175973786041, 0.822747964877635, 0.344312053639442, 0.800653233658522, 0.342199377948418, 0.566993222571909, 0.943046576110646, 0.205751157365739, 0.869258122518659, 0.722073172219098, 0.100285241380334, 0.953388185240328, 0.519595531746745, 0.0752020147629082, 0.681084448704496, 0.993854264030233, 0.332696013618261, 0.647895861417055, 0.931049773702398, 0.227831066353247, 0.339517394546419, 0.418295834213495, 0.7933578514494, 0.230834303423762, 0.36183112137951, 0.112146184314042, 0.909578638384119, 0.463463778840378, 0.641132246702909, 0.438872504513711, 0.249512212816626, 0.461745775071904, 0.95116621744819, 0.889566934667528, 0.158987540984526, 0.297455888940021, 0.813828968908638, 0.621503705624491, 0.197454246226698, 0.159846155671403, 0.323893388034776, 0.296334492508322, 0.516133904224262, 0.504209062084556, 0.0147472680546343, 0.948891679057851, 0.853329840814695, 0.938090464333072, 0.85201144288294, 0.848470917902887, 0.660794902592897, 0.939235628349707, 0.102303135208786, 0.580082803964615, 0.311326453462243, 0.747738091507927, 0.185541106387973, 0.750872075092047, 0.300532962428406, 0.397027931874618, 0.407283964334056, 0.384259346174076, 0.405253793345764, 0.866875231498852, 0.79003885621205, 0.524148418800905, 0.279130646260455, 0.500356029486284, 0.0946899610571563, 0.806238760007545, 0.447147254832089, 0.267661535181105, 0.799683362711221, 0.935482781147584, 0.769920595455915, 0.224836081033573, 0.59492980572395, 0.512453379109502, 0.178810100769624, 0.988842231687158, 0.0841587989125401, 0.942822708981112, 0.237502244766802, 0.593683453975245, 0.0731855148915201, 0.0797033617272973, 0.814499445958063, 0.805948249762878, 0.444702796405181, 0.640953523339704, 0.831606951309368, 0.82150967954658, 0.720124184386805, 0.157033868366852, 0.911784852389246, 0.934505775105208, 0.0158345089294016, 0.837352765724063, 0.643177910707891, 0.440005866810679, 0.738321112003177, 0.0734785106033087, 0.393263341858983, 0.909728143597022, 0.0807785640936345, 0.487831657752395, 0.959145478904247, 0.23206021822989, 0.29854444717057, 0.299084560479969, 0.534687920939177, 0.149291161447763, 0.232058306690305, 0.13559598964639, 0.0185242262668908, 0.265588353853673, 0.649745863629505, 0.813459755387157, 0.570670793065801, 0.583582917228341, 0.801220375578851, 0.478777980431914, 0.207158684032038, 0.394279082305729, 0.0596063644625247, 0.678623552666977, 0.9350078208372, 0.478331675054505, 0.402177171083167, 0.563714587129653, 0.0028468135278672, 0.119647511979565, 0.292315141065046, 0.289680832531303, 0.739024191629142, 0.736178890103474, 0.364605282898992, 0.856575864367187, 0.925258345203474, 0.751978011801839, 0.514282068936154, 0.529927912866697, 0.709864491363987, 0.26476324768737, 0.283819886157289, 0.247430313611403, 0.629851328209043, 0.249849022133276, 0.0598443425260484, 0.562182335415855, 0.0303609946276993, 0.976493626832962, 0.451237421482801, 0.526369270635769, 0.486695694038644, 0.881479129893705, 0.820442848838866, 0.637704129563645, 0.244225610978901, 0.557829858968034, 0.806261986494064, 0.673666044604033, 0.712419882882386, 0.878094075014815, 0.941631552297622, 0.739382346393541, 0.237829465419054, 0.369728234130889, 0.270867840154096, 0.707962230313569, 0.892958819167688, 0.974190018605441, 0.55807022866793, 0.0611096294596791, 0.765021940460429, 0.340049544116482, 0.369983955984935, 0.695213017053902, 0.505749048665166, 0.424490054370835, 0.538598913932219, 0.0545314722694457, 0.126474161399528, 0.207862871931866, 0.717392956605181, 0.447129289386794, 0.435563893057406, 0.191043191822246, 0.842289690161124, 0.256180597702041, 0.651440160814673, 0.831812682561576, 0.405820794170722, 0.411706521874294, 0.699824826326221, 0.426102881319821, 0.381151929264888, 0.0897117902059108, 0.193573871860281, 0.856575455050915, 0.751552684232593, 0.79773143841885, 0.326624975306913), actual = c(1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"