Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4758438        0.4683311 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4802860        0.4802860        0.5001406 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.396611041156575, 0.505969134392217, 0.658696331782266, 0.0828784678597003, 0.595410130685195, 0.145701866131276, 0.429241551086307, 0.737888012314215, 0.799463328206912, 0.0871393173001707, 0.967946906108409, 0.359924559248611, 0.662251062691212, 0.499227886321023, 0.498579261125997, 0.102414463879541, 0.35958613990806, 0.258995465701446, 0.142137671122327, 0.511246718931943, 0.675724992761388, 0.919604440685362, 0.179440701613203, 0.455855068517849, 0.793936823494732, 0.359091443242505, 0.700804154155776, 0.812391377519816, 0.881340101128444, 0.73431344027631, 0.842904142104089, 0.23163419845514, 0.374529826687649, 0.871164435986429, 0.459328671218827, 0.482008947292343, 0.732995352707803, 0.844456493621692, 0.914764242712408, 0.608215855201706, 0.0659790132194757, 0.98056280403398, 0.188864106079563, 0.607279800111428, 0.430925245396793, 0.152380388928577, 0.246678398456424, 0.283762285253033, 0.172857536701486, 0.390858026454225, 0.363360926276073, 0.337546437745914, 0.387882766546682, 0.0281907664611936, 0.902435707626864, 0.870023228926584, 0.854482552036643, 0.527699808124453, 0.327043866738677, 0.360027733026072, 0.583641936536878, 0.923366328468546, 0.442300632363185, 0.0725901287514716, 0.885132516268641, 0.125212891958654, 0.107737528160214, 0.95507317292504, 0.283041065325961, 0.204889614135027, 0.472968991380185, 0.157152363331988, 0.956225877860561, 0.210571581497788, 0.974511072039604, 0.736458550207317, 0.855998000362888, 0.767831800272688, 0.810748424381018, 0.412442532135174, 0.0501933391205966, 0.196439599152654, 0.0172577670309693, 0.649770172312856, 0.590382371330634, 0.205293477745727, 0.904968553222716, 0.29873650544323, 0.692924894625321, 0.0802742054220289, 0.731789663666859, 0.39823315734975, 0.757074971450493, 0.62409447110258, 0.363798624137416, 0.427324739051983, 0.502943369327113, 0.281208067666739, 0.891056096646935, 0.410027167992666, 0.830272388877347, 0.0227874417323619, 0.859981374815106, 0.206067064544186, 0.506886203307658, 0.306755022378638, 0.220403610728681, 0.0174824264831841, 0.668058492708951, 0.622649863595143, 0.555579416686669, 0.0645017044153064, 0.051324640866369, 0.939098778646439, 0.549315477488562, 0.821474618976936, 0.397021517856047, 0.961073324084282, 0.647491281153634, 0.965370290214196, 0.981950754765421, 0.545028887689114, 0.666314186528325, 0.909632246475667, 0.683403773233294, 0.759136846521869, 0.215031186351553, 0.520999737549573, 0.013500879984349, 0.151849172078073, 0.37981390254572, 0.207351478049532, 0.830614865059033, 0.850479927379638, 0.83428384247236, 0.411172688705847, 0.247444852022454, 0.00880036735907197, 0.472956340759993, 0.449193567968905, 0.606609254609793, 0.59268502658233, 0.505218098172918, 0.15857792343013, 0.380768717033789, 0.558301680255681, 0.837773951003328, 0.149867627071217, 0.206081735203043, 0.33425113465637, 0.828342044260353, 0.00183141091838479, 0.721114452229813, 0.786399628734216, 0.554459921782836, 0.990814112825319, 0.952668459154665, 0.436343852430582, 0.892048072535545, 0.499018535716459, 0.265274900943041, 0.804445669054985, 0.548257972579449, 0.168235154822469, 0.596368645085022, 0.190539318602532, 0.0456720693036914, 0.35671646380797, 0.663196266628802, 0.647753122262657, 0.741221496835351, 0.125448933569714, 0.337612327653915, 0.505579267861322, 0.833363040583208, 0.232390544610098, 0.655247350689024, 0.877044902415946, 0.524736376944929, 0.791468370240182, 0.0967030751053244, 0.423988006543368, 0.528420723741874, 0.356327509041876, 0.534867871319875, 0.538857424398884, 0.356877258280292, 0.563516046851873, 0.370422330452129, 0.195251288358122, 0.594603288453072, 0.667847839649767, 0.133282284252346, 0.199361068196595, 0.500580458668992, 0.233971587382257, 0.82381043350324, 0.178569918731228, 0.773481609299779, 0.0186345027759671, 0.42273193737492, 0.307457041926682, 0.311100353719667, 0.718889043200761, 0.270467829657719, 0.102054855320603, 0.958247734233737, 0.449529782868922, 0.807525133481249, 0.720147114712745, 0.592087663011625, 0.0944332252256572, 0.259432473219931, 0.895587891340256, 0.805092116817832, 0.509482610970736, 0.959033395862207, 0.544322518864647, 0.0586813378613442, 0.416666670702398, 0.200656655244529, 0.321487715933472, 0.879137183306739, 0.791968126315624, 0.0991057516075671, 0.0996072543784976, 0.685764249647036, 0.0628424517344683, 0.899979441892356, 0.915456250077114, 0.0934980432502925, 0.344290389912203, 0.979413228807971, 0.554068157216534, 0.927894254215062, 0.431961609050632, 0.849669397575781, 0.162738480139524, 0.711487771943212, 0.750860559986904, 0.0309151739347726, 0.402385692112148, 0.976450502406806, 0.0597277528140694, 0.115038609597832, 0.0993813946843147, 0.707545895129442, 0.377984221559018, 0.350554664619267, 0.622452873503789, 0.463160041952506, 0.0869906912557781, 0.637639384018257, 0.011644379934296, 0.59417764027603, 0.860120689263567, 0.841966923559085, 0.70932106138207, 0.227288192370906, 0.420676823472604, 0.577520505525172, 0.91571926837787, 0.218797594076023, 0.636402189964429, 0.0999624140094966, 0.269110822118819, 0.854260403197259, 0.515228824224323, 0.468181788688526, 0.325628033140674, 0.883953989949077, 0.479664067039266, 0.95705499430187, 0.141700454754755, 0.592387492768466, 0.759534908458591, 0.0979840750806034, 0.419669725000858, 0.919664792949334, 0.4552218161989, 0.499120553722605, 0.294925024500117, 0.149931689957157, 0.43250842159614, 0.0490125250071287, 0.341350385686383, 0.591129707172513, 0.6533781462349, 0.853015080094337, 0.30457267863676, 0.195069335633889, 0.203724334947765, 0.837946723215282, 0.922027916414663, 0.617058911593631, 0.77216728636995, 0.286400533746928, 0.778454789426178, 0.746497748652473, 0.169750070897862, 0.366658182814717, 0.544406183296815, 0.538945481879637, 0.457748918328434, 0.55494037293829, 0.256038700928912, 0.821047947742045, 0.701742290752009, 0.701899829786271, 0.793872496346012, 0.692591973347589, 0.932196111185476, 0.183118262095377, 0.684156158007681, 0.235667275963351, 0.415011362405494, 0.8660081573762, 0.336179057369009, 0.190943763358518, 0.527705791871995, 0.0663435698952526, 0.710446594748646, 0.139455588301644, 0.671310254605487, 0.0520795739721507, 0.520876015536487, 0.244301243918017, 0.196115911006927, 0.118354430422187, 0.770740347914398, 0.450741999316961, 0.672347981715575, 0.67935586697422, 0.867667568149045, 0.176072661532089, 0.700779830571264, 0.194147515576333, 0.640856601530686, 0.679300276795402, 0.338455024873838, 0.439775233855471, 0.774674511747435, 0.84385721385479, 0.381716774310917, 0.320858518360183, 0.330026958137751, 0.66298839612864, 0.741136860335246, 0.266686908202246, 0.66037740255706, 0.204563894309103, 0.49714179825969, 0.0344807140063494, 0.960229169111699, 0.217525000218302, 0.654456766322255, 0.34361751191318, 0.247750184033066, 0.652080933563411, 0.223186769988388, 0.908234249567613, 0.128234729636461, 0.792658236576244, 0.133240618277341, 0.50092054810375, 0.968105936190113, 0.430898582329974, 0.0666101917158812, 0.0662749516777694, 0.883763122139499, 0.48455846356228, 0.540199727518484, 0.34323094622232, 0.75577435689047, 0.796255030669272, 0.332664280897006, 0.342098449822515, 0.0774037067312747, 0.269005649024621, 0.440075953723863, 0.107831280678511, 0.355645415838808, 0.93132601166144, 0.241168251028284, 0.619298783596605, 0.535608320264146, 0.0786527565214783, 0.297876759432256, 0.606465228367597, 0.612395581090823, 0.590636385371909, 0.370759370969608, 0.113338312832639, 0.0607664943672717, 0.376917565008625, 0.56583183282055, 0.0323560901451856, 0.458888164488599, 0.407456601271406, 0.866523446049541, 0.817783155245706, 0.488564251922071, 0.231516265310347, 0.733420945471153, 0.236661585280672, 0.428374380106106, 0.183861391386017, 0.888839363120496, 0.012312448117882, 0.656073800753802, 0.391173390904441, 0.955823679454625, 0.0905248115304857, 0.0151538080535829, 0.789106398820877, 0.374518821481615, 0.595772639382631, 0.477425599936396, 0.169743798440322, 0.490947792073712, 0.876335089560598, 0.489921749103814, 0.676162500167266, 0.465860827127472, 0.865180270047858, 0.29851060686633, 0.601370744872838, 0.824195049935952, 0.11171188345179, 0.626804270315915, 0.348800422623754, 0.597871306352317, 0.455820169532672, 0.135171690257266, 0.0419281523209065, 0.790532286744565, 0.583513573743403, 0.945656072814018, 0.409155189990997, 0.73662685113959, 0.192336085950956, 0.852907073451206, 0.195849384646863, 0.798151081893593, 0.692273589782417, 0.876472167205065, 0.113763035740703, 0.542018236592412, 0.908599022077397, 0.951174148824066, 0.27015977492556, 0.00139267952181399, 0.3770462567918, 0.622571892105043, 0.753543301252648, 0.461327320896089, 0.262652869336307, 0.389783330960199, 0.608924258500338, 0.639599525835365, 0.704642621334642, 0.868236757116392, 0.486834676936269, 0.683365955250338, 0.0707322624512017, 0.829958646092564, 0.321159192593768, 0.541238237638026, 0.573096511652693, 0.0400583760347217, 0.533485820284113, 0.96783446171321, 0.323922573123127, 0.880873269401491, 0.354313360061496, 0.959998924285173, 0.735969512490556, 0.89356414671056, 0.299193136859685, 0.532788182143122, 0.516087269876152, 0.110831243917346, 0.397562317550182, 0.443376974202693, 0.611589571926743, 0.191371083259583, 0.935014780145139, 0.555222743656486, 0.100324013736099, 0.248515119077638, 0.570852957200259, 0.692652700468898, 0.475438442081213, 0.9469408178702, 0.983670426765457, 0.351933646481484, 0.272269623586908, 0.899535631295294, 0.601131015690044, 0.268515710020438, 0.796863141702488, 0.428963945247233, 0.638091056840494, 0.465027408208698, 0.667584317037836, 0.0407575229182839, 0.178154164459556, 0.142975306371227, 0.45511364703998, 0.764294715598226, 0.776231541763991, 0.843324023066089, 0.75444165430963, 0.153120499337092, 0.542710823006928, 0.950600855518132, 0.312836900353432, 0.798140408238396, 0.344890719978139, 0.0655990282539278, 0.79511030158028, 0.923571474384516, 0.732089447090402, 0.161386922933161, 0.596739064669237, 0.295553782954812, 0.857703495770693, 0.745591489830986, 0.133147411048412, 0.619771255645901, 0.946533701848239, 0.446163438260555, 0.978297436609864, 0.425417724763975, 0.697984443744645, 0.65503644454293, 0.980007719947025, 0.103912093210965, 0.237788028782234, 0.368279559770599, 0.919687205459923, 0.626287025166675, 0.00334088620729744, 0.466797937173396, 0.00155471637845039, 0.813794272020459, 0.547792625147849, 0.699565692571923, 0.481883036438376, 0.0267133840825409, 0.814918065210804, 0.376961374888197, 0.11108954413794, 0.21864964440465, 0.00322394841350615, 0.487204723758623, 0.950673197163269, 0.450600909767672, 0.368858634959906, 0.64895549044013, 0.134308400796726, 0.524068708531559, 0.500364334089682, 0.230793949682266, 0.222789312945679, 0.256933931959793, 0.679828861728311, 0.763647854793817, 0.0168193571735173, 0.591906732646748, 0.240616287104785, 0.677753732306883, 0.189666215563193, 0.63049884326756, 0.414147800533101, 0.0174633653368801, 0.0316577835474163, 0.244162477785721, 0.879273897269741, 0.706587924389169, 0.109976788982749, 0.143806165317073, 0.365937842288986, 0.179820423247293, 0.337432098807767, 0.0357328781392425, 0.481941773556173, 0.330057355109602, 0.82678541704081, 0.935306368162856, 0.544700981117785, 0.324658941011876, 0.335336431395262, 0.0440513412468135, 0.304852092405781, 0.974583722651005, 0.637072238139808, 0.697033180855215, 0.907497586449608, 0.241607334697619, 0.359859659336507, 0.561584035400301, 0.788005906855687, 0.967637615278363, 0.31610309286043, 0.645749165676534, 0.857421184657142, 0.181983921909705, 0.218990394379944, 0.659688238054514, 0.665100275306031, 0.311206744750962, 0.401335887610912, 0.838033503387123, 0.64221468870528, 0.475287752458826, 0.95993111981079, 0.816677605267614, 0.611410824581981, 0.527484963415191, 0.093052925541997, 0.228019905276597, 0.741014804458246, 0.363642424112186, 0.329093217849731, 0.784955176059157, 0.739446294726804, 0.12242764281109, 0.166929015889764, 0.383049880620092, 0.503938077716157, 0.396687619853765, 0.810735842911527, 0.157658712007105, 0.679148160852492, 0.131622381042689, 0.0816392377018929, 0.0131484810262918, 0.785636824322864, 0.515979004325345, 0.00184734142385423, 0.575115829473361, 0.239305870141834, 0.127652660710737, 0.201811909908429, 0.747888197656721, 0.649091357598081, 0.0140881391707808, 0.81942794052884, 0.720169110456482, 0.375381832476705, 0.705111734336242, 0.675990406656638, 0.694971723947674, 0.524571916088462, 0.589498574147001, 0.20733150257729, 0.991089516552165, 0.297389477025717, 0.929485496599227, 0.305336472345516, 0.315268071368337, 0.462520074797794, 0.289294531336054, 0.92322955885902, 0.27997627761215, 0.967741451691836, 0.881096740718931, 0.469230524031445, 0.353446677094325, 0.759410415543243, 0.952351401792839, 0.842010132968426, 0.276892633643001, 0.54436940443702, 0.128802586114034, 0.391485418425873, 0.140497393207625, 0.557394205592573, 0.194235966308042, 0.198206116678193, 0.802641921676695, 0.461838788585737, 0.113361473428085, 0.408224557992071, 0.95063056377694, 0.6222158302553, 0.041044675046578, 0.61539782397449, 0.102906636893749, 0.974055419908836, 0.636451841099188, 0.605026124510914, 0.128011289052665, 0.211933475919068, 0.847832399653271, 0.559887901414186, 0.232350819278508, 0.276958681410179, 0.939153886167333, 0.914451426360756, 0.172223919769749, 0.304323440417647, 0.852028564084321, 0.387737488374114, 0.587744337273762, 0.0327130488585681, 0.722801458789036, 0.00722368294373155, 0.981506110634655, 0.398109995061532, 0.268171156523749, 0.483290647855029, 0.394514889689162, 0.0606187807861716, 0.921435126336291, 0.588043301831931, 0.0531887619290501, 0.119645271683112, 0.211391318123788, 0.602471439400688, 0.611196880927309, 0.302067899145186, 0.582242348231375, 0.498249751515687, 0.497107754694298, 0.195798783330247, 0.128127980977297, 0.297730672173202, 0.619278249097988, 0.762140331091359, 0.0197520707733929, 0.680727231083438, 0.0461473339237273, 0.02748031495139, 0.557813110295683, 0.0650003999471664, 0.969874093076214, 0.896219456102699, 0.826406176900491, 0.358689360553399, 0.828840044792742, 0.251858323346823, 0.185501264873892, 0.953875578707084, 0.709233148721978, 0.844577039359137, 0.10466258879751, 0.0428387890569866, 0.596500637708232, 0.944069133605808, 0.145262862788513, 0.699413341470063, 0.21950159384869, 0.737362463260069, 0.784643889870495, 0.902058119885623, 0.562431160826236, 0.235883749090135, 0.237868889002129, 0.326754614943638, 0.649111558683217, 0.234422645065933, 0.794074372854084, 0.245249635074288, 0.590976134873927, 0.809522017138079, 0.376354381907731, 0.869443544885144, 0.235328375827521, 0.00129734794609249, 0.0237289930228144, 0.617517925100401, 0.416628235252574, 0.871518867556006, 0.317688801325858, 0.361938014393672, 0.892962678801268, 0.1647067531012, 0.957129367627203, 0.078344915760681, 0.00929315900430083, 0.538202265743166, 0.165253954706714, 0.197630131151527, 0.365162718575448, 0.746670715045184, 0.172581570921466, 0.91637675371021, 0.331870189635083, 0.413952780421823, 0.337743179872632, 0.436183703131974, 0.409411025466397, 0.55387014336884, 0.0207647266797721, 0.864758626790717, 0.779582536313683, 0.323889766121283, 0.0544208188075572, 0.646732003428042, 0.308226683409885, 0.804763603024185, 0.132256815209985, 0.247451985487714, 0.182708114385605, 0.149943006923422, 0.277119837235659, 0.876108958153054, 0.208422512514517, 0.281711334595457, 0.909473015461117, 0.812416759086773, 0.895367488497868, 0.373765517724678, 0.284436122747138, 0.692322497721761, 0.957813584245741, 0.476066249422729, 0.708718988345936, 0.623981182230636, 0.453431631671265, 0.865604013670236, 0.859423416899517, 0.00448834267444909, 0.365704084280878, 0.810344394762069, 0.916447379160672, 0.711429026443511, 0.921466063009575, 0.228771226480603, 0.867234792560339, 0.835868165595457, 0.687988713383675, 0.888994804350659, 0.773283885559067, 0.921753329923376, 0.250057416968048, 0.636075793998316, 0.47129923873581, 0.83358916407451, 0.97163071599789, 0.508241068106145, 0.0641314496751875, 0.141346796648577, 0.832384661072865, 0.08396165096201, 0.0583225714508444, 0.28359288466163, 0.0535931224003434, 0.643767728004605, 0.347581970272586, 0.255448250798509, 0.885487397667021, 0.909206732409075, 0.550446467241272, 0.93045211955905, 0.769934290554374, 0.0270198110956699, 0.354123316705227, 0.57481394498609, 0.462815896607935, 0.913548060692847, 0.571712597506121, 0.216643680818379, 0.445239769760519, 0.159452963620424, 0.459541254211217, 0.404498036019504, 0.511711486848071, 0.642774398671463, 0.123240596614778, 0.130534504540265, 0.153921969234943, 0.347304394934326, 0.0335313458926976, 0.000419131712988019, 0.558423767099157, 0.055968772387132, 0.369196948129684, 0.0960988402366638, 0.078583586262539, 0.632422037888318, 0.416438422165811, 0.277997791301459, 0.896163797471672, 0.759661737130955, 0.879400670994073, 0.81558301881887, 0.978478447534144, 0.633229720639065, 0.552856878144667, 0.552029774524271, 0.71920648124069, 0.752862520283088, 0.317338828928769, 0.48870195238851, 0.580525496276096, 0.902165036881343, 0.792707080952823, 0.0843767013866454, 0.41861638892442, 0.460096332244575, 0.532017751596868, 0.441243905341253, 0.324434428010136, 0.74021547823213, 0.635124762076885, 0.132927090860903, 0.0764825772494078, 0.442883474985138, 0.714173488551751, 0.242636510869488, 0.410908399848267, 0.27370405732654, 0.391973795834929, 0.81019460526295, 0.825013539288193, 0.388720708200708, 0.166255521820858, 0.192936934065074, 0.510234539629892, 0.755116594489664, 0.237507647601888, 0.449729148065671, 0.897339351242408, 0.335700596915558, 0.491084365639836, 0.251071201870218, 0.429447127971798, 0.809764434816316, 0.494285825872794, 0.861742960754782, 0.17528489115648, 0.146946954075247, 0.54810459446162, 0.00938930734992027, 0.875005199806765, 0.23386368737556, 0.981071128044277, 0.496113028144464, 0.834751267218962, 0.176717161666602, 0.10674896184355, 0.808243364561349, 0.156850370578468, 0.139577077468857, 0.0909981967415661, 0.199227578472346, 0.948827551444992, 0.295091854874045, 0.54490161780268, 0.197831262368709, 0.66919441265054, 0.661112012108788, 0.257103624986485, 0.409990339539945, 0.139906966360286, 0.861149697331712, 0.177001884207129, 0.713746966794133, 0.706120525486767, 0.921227670973167, 0.178426629398018, 0.275164897786453, 0.120316843967885, 0.747703403700143, 0.887343886774033, 0.549900442827493, 0.389362409478053, 0.240230002673343, 0.601728049805388, 0.947318224003538, 0.794838840141892, 0.415898304898292, 0.648413478862494, 0.563274742802605, 0.901542387669906, 0.0360493876505643, 0.249290818348527, 0.492561691440642, 0.241274718660861, 0.853701853426173, 0.140298493904993, 0.161618456244469, 0.429702485213056, 0.0718391537666321, 0.204771675635129, 0.62699426477775, 0.570953494636342, 0.929208994843066, 0.865223094355315, 0.70221068081446, 0.582530020270497, 0.866239274851978, 0.2110294487793, 0.369932467583567, 0.100209518801421, 0.841811645543203, 0.18408944667317), actual = c(1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.753111309371889, 0.426955783739686, 0.886266660410911, 0.542354721110314, 0.120731743285432, 0.696276759728789, 0.413429680513218, 0.23223391501233, 0.580576977692544, 0.743332503130659, 0.270474427845329, 0.151696973945946, 0.210692339809611, 0.624426109716296, 0.269250140991062, 0.373595011886209, 0.358468370279297, 0.0141798011027277, 0.484166555339471, 0.31155279930681, 0.606781027046964, 0.265467431861907, 0.197744659846649, 0.827258562436327, 0.846292253118008, 0.0899460497312248, 0.126113407313824, 0.370590540580451, 0.53636338096112, 0.397662077331915, 0.554434473859146, 0.135241969488561, 0.373630827991292, 0.141850732732564, 0.0129547668620944, 0.695450210012496, 0.465177545323968, 0.96068599796854, 0.762290904065594, 0.480410910909995, 0.0794812147505581, 0.0350663890130818, 0.284191936254501, 0.859791457187384, 0.581690369406715, 0.673602275084704, 0.506328780436888, 0.190883120754734, 0.207038959255442, 0.170436095446348, 0.406942927045748, 0.548887266078964, 0.563867697259411, 0.300369844539091, 0.130858489545062, 0.205477887764573, 0.146802455186844, 0.815316364634782, 0.491845088312402, 0.363726952578872, 0.797086354577914, 0.10144779086113, 0.847968841670081, 0.452019618358463, 0.705436627846211, 0.434451600303873, 0.257466302253306, 0.887643772410229, 0.602713758358732, 0.808383995201439, 0.935328613966703, 0.547048019245267, 0.213707619346678, 0.128069904632866, 0.83655092632398, 0.170539660844952, 0.240791833028197, 0.965388159034774, 0.00620329193770885, 0.299691292690113, 0.247234497917816, 0.329974572407082, 0.301788252778351, 0.899536757031456, 0.316795316059142, 0.853382703382522, 0.548884743824601, 0.489702421007678, 0.915474578505382, 0.313257201807573, 0.877179905306548, 0.256455697352067, 0.3496659679804, 0.0109492358751595, 0.92875236296095, 0.810304120182991, 0.229537894017994, 0.622238992713392, 0.102914737304673, 0.481987397652119, 0.384367554215714, 0.404069124953821, 0.430986242834479, 0.225518316728994, 0.305626325309277, 0.305465438403189, 0.519437328679487, 0.489846222801134, 0.612293393816799, 0.782230884768069, 0.296835748013109, 0.507111892104149, 0.0417271631304175, 0.70267377817072, 0.77517133555375, 0.948010767577216, 0.764760515652597, 0.0921601234003901, 0.997228292981163, 0.712635493604466, 0.748306484660134, 0.113956260727718, 0.720465518767014, 0.968604220775887, 0.473556093871593, 0.693897425895557, 0.601737757446244, 0.165545720141381, 0.646553051425144, 0.515815901337191, 0.744138343259692, 0.501931089442223, 0.638171136612073, 0.876429608790204, 0.670939564239234, 0.0181666542775929, 0.654423014260828, 0.0373071348294616, 0.355829359730706, 0.219250824768096, 0.940091698430479, 0.278707208577543, 0.786129979416728, 0.526319748256356, 0.48770463373512, 0.624804664868861, 0.522269481560215, 0.190797158051282, 0.608164369128644, 0.381024670321494, 0.24822073802352, 0.183475178666413, 0.847529097925872, 0.138049018802121, 0.106455616420135, 0.57252238295041, 0.755676982458681, 0.323477909900248, 0.57833749987185, 0.786978555610403, 0.378653750289232, 0.457106775371358, 0.344771432224661, 0.897134334314615, 0.990685869241133, 0.162478718208149, 0.149886040715501, 0.0777581126894802, 0.814571852097288, 0.551856358069927, 0.268841036129743, 0.357588182669133, 0.146908485330641, 0.455219916533679, 0.982962525915354, 0.622327459044755, 0.530223079258576, 0.0650211917236447, 0.643360721180215, 0.514771567191929, 0.868512168992311, 0.761428378522396, 0.349313073791564, 0.0553237453568727, 0.031833030981943, 0.271753839449957, 0.0507153868675232, 0.873407974373549, 0.676710067084059, 0.392090521752834, 0.0341083763632923, 0.082981871208176, 0.13937900820747, 0.641397284576669, 0.931978330248967, 0.420718523906544, 0.23939130268991, 0.201950060203671, 0.234982082154602, 0.279994347831234, 0.904334648279473, 0.488661215640604, 0.337626857450232, 0.631225317250937, 0.992276510456577, 0.193895776756108, 0.954661003546789, 0.164284768048674, 0.916917006950825, 0.500501825241372, 0.0745006413199008, 0.677520510740578, 0.772537431214005, 0.467283423990011, 0.742030302528292, 0.0892264156136662, 0.855125074042007, 0.25431782216765, 0.399687124416232, 0.387250771513209, 0.704976642504334, 0.258719560923055, 0.257781690452248, 0.735019294312224, 0.207836288725957, 0.007960265269503, 0.756000594934449, 0.63583159330301, 0.748185029253364, 0.432288771960884, 0.0744327278807759, 0.452746408060193, 0.189647246617824, 0.354549394687638, 0.867944197263569, 0.387083651265129, 0.609845089027658, 0.849563559750095, 0.364866836927831, 0.326861002715304, 0.171934793470427, 0.429642082890496, 0.372791573172435, 0.581560743972659, 0.210282222600654, 0.742040599230677, 0.968687835847959, 0.690942778717726, 0.693576862104237, 0.160300598945469, 0.0612059973645955, 0.208196519874036, 0.822538363281637, 0.746259142179042, 0.983599886065349, 0.948012468637899, 0.166288636391982, 0.622973300749436, 0.545423872536048, 0.577959035057575, 0.917374707292765, 0.306141706183553, 0.0965105199720711, 0.826457537012175, 0.44056759448722, 0.507602393394336, 0.233559022191912, 0.567582451971248, 0.816199665889144, 0.218983282567933, 0.989207210484892, 0.217246802756563, 0.88025560346432, 0.236034867120907, 0.938328989082947, 0.914114638464525, 0.388098625931889, 0.426053362200037, 0.888403868768364, 0.314658461837098, 0.952424164395779, 0.263472062302753, 0.561215680092573, 0.251945142867044, 0.145056567154825, 0.179082859540358, 0.998415217036381, 0.842907880200073, 0.645378873683512, 0.532514120219275, 0.261198967229575, 0.358125362079591, 0.542708126595244, 0.551387265091762, 0.891445158282295, 0.23482986423187, 0.98149313707836, 0.782285226974636, 0.964086720254272, 0.000307806534692645, 0.0132386025507003, 0.858556215418503, 0.219213948585093, 0.855456526158378, 0.67101002833806, 0.0773957401979715, 0.242176707368344, 0.075548552442342, 0.524133894825354, 0.190367663046345, 0.539071781095117, 0.553983404068276, 0.810606800019741, 0.646230530226603, 0.055133156478405, 0.300545804901049, 0.202899700729176, 0.600499526597559, 0.450614800909534, 0.687673972919583, 0.0351102920249104, 0.871766331372783, 0.695913642644882, 0.36884240154177, 0.175408779177815, 0.444938522065058, 0.548841228242964, 0.772344477009028, 0.106259088497609, 0.146225565578789, 0.333449440775439, 0.194991595810279, 0.0758859231136739, 0.481530228629708, 0.743877538014203, 0.439619606127962, 0.851541214622557, 0.452317490242422, 0.126909692538902, 0.0400462704710662, 0.630644130287692, 0.80650344141759, 0.597145967651159, 0.0862006167881191, 0.68465695832856, 0.908439997350797, 0.165839767316356, 0.241438763681799, 0.0528327543288469, 0.167381652398035, 0.241036963183433, 0.590150585398078, 0.583610532106832, 0.331981252878904, 0.171295954845846, 0.294395548291504, 0.539948726538569, 0.33875881251879, 0.725506422575563, 0.527365887770429, 0.485101060708985, 0.745587545912713, 0.801477739587426, 0.899940795265138, 0.416004579514265, 0.283927674405277, 0.595020582200959, 0.326601690612733, 0.00895106582902372, 0.954457409214228, 0.665307628922164, 0.620261758100241, 0.956815309356898, 0.0439811346586794, 0.930235211737454, 0.611835613846779, 0.0847438105847687, 0.885129156988114, 0.329569674562663, 0.70685385260731, 0.0658723113592714, 0.16376829543151, 0.678910573245957, 0.256687378976494, 0.135345384012908, 0.495686729205772, 0.633222135016695, 0.076057915808633, 0.937047231476754, 0.733921116916463, 0.678249506512657, 0.255807663081214, 0.838955779094249, 0.753171676304191, 0.714954071445391, 0.574276468250901, 0.956370478495955, 0.437834894284606, 0.434203859651461, 0.38603982469067, 0.986226864624768, 0.361179173691198, 0.467191172996536, 0.229696701280773, 0.372188826324418, 0.372311809565872, 0.554780600126833, 0.572922097053379, 0.678445992991328, 0.843383841682225, 0.305603012442589, 0.026196621125564, 0.594751936150715, 0.00639449874870479, 0.790789800928906, 0.756499938666821, 0.455421146471053, 0.909595015691593, 0.793466369155794, 0.866474291309714, 0.338259622221813, 0.407279167789966, 0.639377466170117, 0.590579090407118, 0.810631481930614, 0.990150745492429, 0.651032972382382, 0.420956058194861, 0.46249248390086, 0.0427581872791052, 0.251260238699615, 0.221615288639441, 0.298320067580789, 0.939509337535128, 0.147508375812322, 0.295627238461748, 0.0820467160083354, 0.258777967886999, 0.274115119827911, 0.770630250684917, 0.142910729395226, 0.0320934078190476, 0.759117812151089, 0.760212678229436, 0.459050290752202, 0.37808828195557, 0.142034448916093, 0.0622371796052903, 0.882438836852089, 0.198991113575175, 0.329214355209842, 0.260571649298072, 0.631941033760086, 0.751920843031257, 0.830033339560032, 0.0179395515006036, 0.158353850245476, 0.309547875309363, 0.142107222694904, 0.908637608168647, 0.876392300706357, 0.613523582462221, 0.363265746738762, 0.797266212059185, 0.0836760431993753, 0.679836039198563, 0.413545276736841, 0.19027702184394, 0.596796744503081, 0.767951304325834, 0.918917004950345, 0.847999651217833, 0.665471703512594, 0.119886630680412, 0.684921780135483, 0.669979290338233, 0.414617978734896, 0.886763638351113, 0.305394130060449, 0.731655600713566, 0.536018246551976, 0.952887675957754, 0.776153987273574, 0.752078419551253, 0.664990476099774, 0.227207499323413, 0.309335155179724, 0.860326565802097, 0.961959646316245, 0.0467120734974742, 0.379452615743503, 0.880731895798817, 0.812450372381136, 0.276354736182839, 0.724538456648588, 0.642139136791229, 0.223291539121419, 0.0456920003052801, 0.231387894367799, 0.974920934531838, 0.0357394877355546, 0.959357284940779, 0.83001904678531, 0.285540010547265, 0.666069282451645, 0.806546870619059, 0.670065682148561, 0.491566632641479, 0.956923387944698, 0.712854207726195, 0.529324213275686, 0.415612095734105, 0.704859375488013, 0.0489884249400347, 0.970550834666938, 0.571736170211807, 0.633082176093012, 0.237406380241737, 0.979082149453461, 0.716465085046366, 0.707229946972802, 0.314091864973307, 0.221484858542681, 0.0198659040033817, 0.471776365069672, 0.620416215155274, 0.108839468332008, 0.159529010532424, 0.0279418386053294, 0.423473251052201, 0.899735807441175, 0.340547845466062, 0.532318401848897, 0.468016313854605, 0.271036769961938, 0.647301560500637, 0.801607557339594, 0.0273010416422039, 0.452094811946154, 0.478209117194638, 0.395716782193631, 0.644650359405205, 0.077001056401059, 0.674050119239837, 0.339041563915089, 0.287586601218209, 0.392958419863135, 0.769339153543115, 0.165202936390415, 0.876804104773328, 0.327266992302611, 0.248637827811763, 0.906722868327051, 0.0428120808210224, 0.261776940431446, 0.306796912802383, 0.51849906751886, 0.421068809460849, 0.949281660141423, 0.597080630250275, 0.157723419368267, 0.645403356291354, 0.823515802621841, 0.175761567661539, 0.0187368150800467, 0.682711858069524, 0.580447034910321, 0.819504429819062, 0.240907510509714, 0.254604123067111, 0.804658623645082, 0.0968908020295203, 0.825611088424921, 0.0421425821259618, 0.749129920266569, 0.641531503060833, 0.595555542036891, 0.0973657213617116, 0.75610677478835, 0.0566404613200575, 0.293389124562964, 0.398556679021567, 0.0611642634030432, 0.673564000288025, 0.000886332942172885, 0.131231546401978, 0.859632302541286, 0.724519215058535, 0.737710972549394, 0.391163373365998, 0.975313768256456, 0.374417666345835, 0.374587714672089, 0.200573332607746, 0.499308404978365, 0.957536989590153, 0.00556503795087337, 0.283327652607113, 0.877853964921087, 0.118410389171913, 0.997522507095709, 0.822335608769208, 0.389073977945372, 0.571730540134013, 0.477439288981259, 0.255575907183811, 0.990143280709162, 0.938316159183159, 0.617605709237978, 0.139160403748974, 0.510361657245085, 0.783446778776124, 0.280619849916548, 0.724823234137148, 0.29424648382701, 0.565194244962186, 0.760208370862529, 0.58569532353431, 0.395219132769853, 0.285806220024824, 0.161031260853633, 0.155880784383044, 0.941418138798326, 0.788571001030505, 0.0501661573071033, 0.889504168182611, 0.554496987257153, 0.0674005870241672, 0.15525488415733, 0.487621744628996, 0.558706029783934, 0.121986594051123, 0.0663362874183804, 0.832477416144684, 0.721522400155663, 0.560855759540573, 0.411312677897513, 0.347045126371086, 0.537757513811812, 0.816262708278373, 0.760544616961852, 0.795301865786314, 0.275306560564786, 0.460148357553408, 0.259442159440368, 0.247785377316177, 0.958561245119199, 0.189939816482365, 0.597798811038956, 0.720030610449612, 0.674005256034434, 0.66557297273539, 0.175481079611927, 0.430377896409482, 0.817658625077456, 0.909256895305589, 0.467000915668905, 0.181180333020166, 0.456122904317454, 0.407483088318259, 0.948718190658838, 0.66874540806748, 0.542394147254527, 0.277835888788104, 0.612387198256329, 0.421351760160178, 0.81219139508903, 0.50652237306349, 0.54654406523332, 0.974361022002995, 0.428114159265533, 0.15812905644998, 0.519606909016147, 0.399699186440557, 0.0180142377503216, 0.234243103535846, 0.880527777364478, 0.644759667571634, 0.656288508791476, 0.351897030835971, 0.0298984458204359, 0.980316153494641, 0.481341920793056, 0.0341969232540578, 0.801547118462622, 0.379635520046577, 0.418463533744216, 0.0971819024998695, 0.0771422111429274, 0.964764025295153, 0.235358800040558, 0.584144460037351, 0.0868403005879372, 0.542637474136427, 0.678829595912248, 0.317184400744736, 0.022335589164868, 0.270126855932176, 0.604839237639681, 0.0372685077600181, 0.590925463242456, 0.40046585095115, 0.850091357482597, 0.393503631465137, 0.350675104884431, 0.894931748276576, 0.584447370143607, 0.625920919002965, 0.256337044760585, 0.132752472069114, 0.629420607350767, 0.74873834149912, 0.865892438683659, 0.952679547714069, 0.104645382612944, 0.545417977496982, 0.998714885441586, 0.956592812668532, 0.0544839608483016, 0.924728619167581, 0.585228403564543, 0.820174050051719, 0.274258743505925, 0.974045876180753, 0.443593895295635, 0.472235664492473, 0.645954122534022, 0.479831924894825, 0.213453886564821, 0.549688477767631, 0.708839277038351, 0.798529093619436, 0.804215845884755, 0.716027504298836, 0.59196489257738, 0.856566699687392, 0.554700837936252, 0.120198756689206, 0.550556430593133, 0.305193572305143, 0.852734330808744, 0.502757404465228, 0.53114302828908, 0.994309767382219, 0.789732688805088, 0.926751883001998, 0.0425709011033177, 0.745216904440895, 0.491474135778844, 0.0190177953336388, 0.436730117537081, 0.949060564627871, 0.313879003049806, 0.308626212412491, 0.526374403154477, 0.232266142033041, 0.428872969700024, 0.00523568573407829, 0.561482087243348, 0.464498771121725, 0.122387346578762, 0.0220721936784685, 0.614443291677162, 0.504265964031219, 0.780903226463124, 0.962300582556054, 0.867690037703142, 0.153523549903184, 0.146987446583807, 0.572751914151013, 0.280853219795972, 0.900922288652509, 0.978325701551512, 0.984694371232763, 0.579360025934875, 0.253836241550744, 0.446986447321251, 0.736379584763199, 0.715230384841561, 0.881937625119463, 0.126331319101155, 0.110947328153998, 0.0153175885789096, 0.013512765057385, 0.471322559285909, 0.800435506505892, 0.894663124578074, 0.675675845937803, 0.847792294807732, 0.943358219461516, 0.372796165524051, 0.189815409947187, 0.670490195509046, 0.908504964783788, 0.527717787306756, 0.725004068342969, 0.925103460904211, 0.433510240865871, 0.597414395771921, 0.211970059201121, 0.653010791866109, 0.684300866909325, 0.387906646123156, 0.242668688297272, 0.643625861499459, 0.835764549905434, 0.23111636354588, 0.55235452670604, 0.766946926480159, 0.824990910943598, 0.753858862677589, 0.266598025104031, 0.961724377935752, 0.98718559066765, 0.329336693277583, 0.0762300740461797, 0.518363710027188, 0.972563321702182, 0.697730524465442, 0.265878027305007, 0.821057254448533, 0.28795998939313, 0.0657786568626761, 0.593724360689521, 0.122285875026137, 0.787884595571086, 0.253546418389305, 0.262959301006049, 0.0949350164737552, 0.088641882641241, 0.282440218375996, 0.993913072859868, 0.810647463658825, 0.844060280593112, 0.951381466118619, 0.643216556403786, 0.400521227624267, 0.687783734174445, 0.735465307021514, 0.244384978665039, 0.687011899892241, 0.120844580465928, 0.40764604927972, 0.134449885226786, 0.47555558825843, 0.372791523346677, 0.141561553347856, 0.489360773935914, 0.146294292295352, 0.583295522490516, 0.75138402194716, 0.595078038051724, 0.003177335485816, 0.614109015325084, 0.641061632428318, 0.933543969411403, 0.370448344387114, 0.79627222684212, 0.577861602650955, 0.223501513944939, 0.108645192580298, 0.010215375572443, 0.975070585729554, 0.102506214054301, 0.574937063734978, 0.699513592291623, 0.41999045596458, 0.0773030440323055, 0.596297156997025, 0.873115221969783, 0.222039976622909, 0.432817987399176, 0.260759198339656, 0.00141462171450257, 0.261215990874916, 0.221390176564455, 0.4022097280249, 0.18125038431026, 0.676100654061884, 0.77889012824744, 0.257085362216458, 0.362782449927181, 0.167771273758262, 0.995782908285037, 0.937846991699189, 0.850957410177216, 0.551506811985746, 0.416198661318049, 0.422391558997333, 0.379864221438766, 0.298282754840329, 0.29972273320891, 0.377574235433713, 0.511161394184455, 0.672799650812522, 0.496131253661588, 0.252992511494085, 0.689563550287858, 0.438970761140808, 0.667982849758118, 0.00380614260211587, 0.61199815175496, 0.660189553396776, 0.133989718509838, 0.895857473602518, 0.344285322120413, 0.945075214607641, 0.811209300067276, 0.253771275514737, 0.780001734849066, 0.103892655111849, 0.887354382779449, 0.61520968307741, 0.231763281160966, 0.240919106639922, 0.0775808834005147, 0.498233055463061, 0.913336230907589, 0.326930606737733, 0.426540121203288, 0.141504063270986, 0.515989704988897, 0.648132851580158, 0.422468569595367, 0.886438166489825, 0.604310034075752, 0.780992281623185, 0.679164358181879, 0.198721585329622, 0.0322810348588973, 0.118865015450865, 0.854357878211886, 0.689338611904532, 0.021220664260909, 0.602177403401583, 0.632152239326388, 0.808782377745956, 0.73403197247535, 0.656491572968662, 0.47908575902693, 0.128536870935932, 0.54942315351218, 0.348741673864424, 0.431755823316053, 0.839278703788295, 0.146048057358712, 0.663212704006582, 0.911621472565457, 0.115800783503801, 0.512491379864514, 0.586039969930425, 0.240038327407092, 0.587540405569598, 0.474024448776618, 0.876985921990126, 0.699697432573885, 0.0886751944199204, 0.319694406585768, 0.711238569114357, 0.925175440963358, 0.995191981550306, 0.940290314843878, 0.61011921800673, 0.804430765565485, 0.864372825017199, 0.516006899997592, 0.942869949620217, 0.919262602459639, 0.756952736759558, 0.278509182389826, 0.307619419647381, 0.40748046641238, 0.223944341065362, 0.901695109438151, 0.470146361738443, 0.860691473353654, 0.809643373126164, 0.569591963198036, 0.406909335637465, 0.974144138395786, 0.0532477919477969, 0.300880997208878, 0.991292686201632, 0.902004112489522, 0.545507474569604, 0.596869060304016, 0.706497114384547, 0.825190035393462, 0.99068519892171, 0.981480254093185, 0.387210725108162, 0.0308926478028297, 0.356193418148905, 0.752863909117877, 0.291336155263707, 0.184445699909702, 0.92739969282411, 0.205004307208583, 0.358918320853263, 0.333384033991024), actual = c(0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"