Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.3158626        0.3157886 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4334629        0.4334629        0.5026993 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.751788361230865, 0.498703153803945, 0.995572068961337, 0.205070903757587, 0.846426655072719, 0.882520945975557, 0.139383231988177, 0.435230202507228, 0.491798686329275, 0.119648955063894, 0.980221955571324, 0.166459622094408, 0.481150205712765, 0.308130596531555, 0.326258223270997, 0.826396206161007, 0.996014597360045, 0.93961097416468, 0.61831170739606, 0.176676280563697, 0.787845449754968, 0.780719928909093, 0.897923862794414, 0.646387574030086, 0.873682454926893, 0.528242975240573, 0.0316063924692571, 0.74045186303556, 0.921829639701173, 0.579487712820992, 0.428562001092359, 0.624997867969796, 0.145043937955052, 0.384606350678951, 0.14684973587282, 0.434909506933764, 0.323253470472991, 0.623637405456975, 0.894175603520125, 0.426578635117039, 0.095931560266763, 0.591450028121471, 0.42515202681534, 0.28020112356171, 0.247931912075728, 0.152794679626822, 0.767900503939018, 0.581525323214009, 0.594394715270028, 0.943364780163392, 0.495346086798236, 0.775993576971814, 0.470357349142432, 0.58426978578791, 0.912041034782305, 0.213077273685485, 0.362690704409033, 0.791066044010222, 0.574119507567957, 0.996860435698181, 0.121619164943695, 0.239460754208267, 0.213277762522921, 0.516601907322183, 0.949247339973226, 0.535154475597665, 0.000208607409149408, 0.384924467187375, 0.453621057560667, 0.049074197653681, 0.890228126896545, 0.301030814182013, 0.0275703778024763, 0.469821102684364, 0.856742113595828, 0.776572618633509, 0.840939316200092, 0.663595012389123, 0.235163040691987, 0.578006633091718, 0.874521564925089, 0.157435107044876, 0.672361767385155, 0.26008298411034, 0.652550043771043, 0.488578656688333, 0.684768239269033, 0.278356381924823, 0.75362564297393, 0.677944721421227, 0.292669308837503, 0.914413394872099, 0.67183409514837, 0.612494095228612, 0.336636312073097, 0.491358712315559, 0.792767799925059, 0.902028129668906, 0.377219274407253, 0.851047507487237, 0.0298504340462387, 0.054108721204102, 0.496634704526514, 0.34694816195406, 0.835473847808316, 0.010727122426033, 0.676663449266925, 0.436187775805593, 0.35869138664566, 0.391407183604315, 0.829289232613519, 0.622321549104527, 0.353770004818216, 0.685382266528904, 0.621782639529556, 0.387565834913403, 0.0328089110553265, 0.789959932444617, 0.56428232579492, 0.13691926933825, 0.110708642750978, 0.562612507957965, 0.414998228428885, 0.0929510213900357, 0.432392270304263, 0.894933409057558, 0.53385800588876, 0.242986998287961, 0.968430721201003, 0.978951283497736, 0.371067286236212, 0.656130609801039, 0.754803742282093, 0.720977060962468, 0.329880997072905, 0.759251475334167, 0.208009277703241, 0.137246555648744, 0.898672806099057, 0.128605461213738, 0.477544389665127, 0.196037383750081, 0.694835909642279, 0.707569821737707, 0.511763937072828, 0.446130137424916, 0.211638408713043, 0.504309570183977, 0.0520563449244946, 0.860009991098195, 0.0771876373328269, 0.0778303751721978, 0.816489370306954, 0.743596237152815, 0.726857936708257, 0.142549799056724, 0.302542284829542, 0.0214394484646618, 0.624169662827626, 0.937000905163586, 0.0078930442687124, 0.284314334159717, 0.978706795256585, 0.0785223122220486, 0.418503022752702, 0.1940396532882, 0.967140283901244, 0.783843301702291, 0.454076552065089, 0.565000514499843, 0.688346408307552, 0.314273030497134, 0.301958326250315, 0.409105635015294, 0.756472540553659, 0.5634825960733, 0.718905614688993, 0.456956793321297, 0.00535334111191332, 0.616956542246044, 0.807753189234063, 0.702312448527664, 0.195306495763361, 0.445190257625654, 0.601521087810397, 0.248098091688007, 0.115517369005829, 0.279942821711302, 0.129570780089125, 0.582805077312514, 0.81589774065651, 0.196686320705339, 0.207224085694179, 0.366655606543645, 0.0877785063348711, 0.154292203951627, 0.580053165787831, 0.553737401263788, 0.345564102288336, 0.157251803670079, 0.0965775670483708, 0.295177226420492, 0.435508935013786, 0.843996834475547, 0.297549000475556, 0.273314052494243, 0.633594402810559, 0.494052985450253, 0.356276310747489, 0.618425951106474, 0.450118824839592, 0.130744494032115, 0.554777909303084, 0.271805339260027, 0.507903120014817, 0.0954965825658292, 0.324324268149212, 0.654146247543395, 0.556156639708206, 0.681780124315992, 0.975031364243478, 0.90586951863952, 0.503396871732548, 0.277737255208194, 0.333291786257178, 0.413104647770524, 0.464737530099228, 0.0540292961522937, 0.650127971777692, 0.79932763450779, 0.074180873343721, 0.327871524030343, 0.98009561561048, 0.754292065743357, 0.449053537100554, 0.603392304386944, 0.852155686123297, 0.534499112516642, 0.272573108784854, 0.845844099996611, 0.468484920216724, 0.25426464388147, 0.635626750066876, 0.324259238550439, 0.424951669294387, 0.465017607668415, 0.619706586468965, 0.531330343801528, 0.395321745658293, 0.463849808322266, 0.964664049213752, 0.122408602386713, 0.438050082651898, 0.574610898736864, 0.442000878974795, 0.939905635081232, 0.435733873164281, 0.740195960504934, 0.0319622389506549, 0.394405908649787, 0.655676765134558, 0.733417760580778, 0.126689797034487, 0.180818248772994, 0.653917303774506, 0.250244666822255, 0.595623418688774, 0.500844955909997, 0.795018072472885, 0.602726965211332, 0.746402199147269, 0.930159560171887, 0.0845410865731537, 0.839493052568287, 0.785036960849538, 0.435420555295423, 0.190571793122217, 0.0735696828924119, 0.556246657855809, 0.162400255911052, 0.173965942114592, 0.666464652167633, 0.811612474964932, 0.209823047043756, 0.418999313609675, 0.00483960146084428, 0.683894444257021, 0.556992066791281, 0.921242700191215, 0.346274667419493, 0.258672049501911, 0.994245553854853, 0.256989093031734, 0.168693494983017, 0.00979794911108911, 0.686586454743519, 0.19214606564492, 0.549815032631159, 0.882887140847743, 0.322157397866249, 0.197311531053856, 0.0737501305993646, 0.144611021736637, 0.428961751749739, 0.505737959640101, 0.968564964830875, 0.964006811147556, 0.366129784844816, 0.436141504207626, 0.595809646183625, 0.952123600523919, 0.901978579116985, 0.714408909203485, 0.773601072141901, 0.937128717312589, 0.00789381214417517, 0.300399600528181, 0.187842662679031, 0.065173870883882, 0.749614250380546, 0.331971906358376, 0.538108527427539, 0.318805363960564, 0.533487402135506, 0.957768908236176, 0.519805366406217, 0.425368805648759, 0.7786287907511, 0.469399120425805, 0.599213457666337, 0.824026979738846, 0.619663068791851, 0.429237165488303, 0.238875390728936, 0.31732562975958, 0.232082019560039, 0.587416235124692, 0.0291412989608943, 0.664434155682102, 0.720027402043343, 0.647842919453979, 0.0509070521220565, 0.298848532838747, 0.239854863611981, 0.279658296843991, 0.996125339763239, 0.56926227780059, 0.595441275043413, 0.883326532552019, 0.400880352128297, 0.193892984418198, 0.19104897370562, 0.860149479936808, 0.279058114858344, 0.687220945721492, 0.629051244584844, 0.46046959143132, 0.62895160028711, 0.822031489107758, 0.639154260279611, 0.559020781191066, 0.811380989616737, 0.453017403837293, 0.155590703012422, 0.864185740472749, 0.41395545611158, 0.801103020552546, 0.497691520722583, 0.978440006496385, 0.214902959764004, 0.0150389012414962, 0.526837966637686, 0.437871660804376, 0.440673132659867, 0.988247133325785, 0.178520956775174, 0.528772140387446, 0.113449830561876, 0.113133061910048, 0.160804571351036, 0.596174077363685, 0.297976990463212, 0.708154576132074, 0.518926920834929, 0.0565359143074602, 0.770168456016108, 0.316306481836364, 0.135417679790407, 0.452637294307351, 0.66696664178744, 0.461261834250763, 0.0648161182180047, 0.434735958930105, 0.591245967196301, 0.306962974835187, 0.186422042548656, 0.588066020514816, 0.406859911512583, 0.846098961075768, 0.196415137732401, 0.82422713842243, 0.364352556644008, 0.00545321288518608, 0.307924552820623, 0.273312078090385, 0.0464201811701059, 0.693488706136122, 0.492652605287731, 0.694877000292763, 0.432857518550009, 0.275991444243118, 0.715440093073994, 0.248446467798203, 0.737044743029401, 0.0313567013945431, 0.14763010549359, 0.970646158559248, 0.473065516678616, 0.534902957733721, 0.332185011124238, 0.380480236839503, 0.788399677956477, 0.464378325501457, 0.93318955367431, 0.11342364246957, 0.155288415960968, 0.923366475384682, 0.434200761374086, 0.0298832610715181, 0.715593996690586, 0.8520522788167, 0.705632539000362, 0.710955887800083, 0.063830798259005, 0.85140922665596, 0.965886964462698, 0.640476723201573, 0.0544298698659986, 0.122521611163393, 0.791181035107002, 0.941108563682064, 0.219927008729428, 0.676749931648374, 0.371742375427857, 0.316025295061991, 0.545162143651396, 0.513023102656007, 0.960787352873012, 0.551916340133175, 0.317752206232399, 0.466425776015967, 0.422409631311893, 0.764587206766009, 0.120805985061452, 0.922643946483731, 0.414248085115105, 0.639494985574856, 0.0884625960607082, 0.619176649488509, 0.0172130721621215, 0.784053759183735, 0.0444059746805578, 0.93642876110971, 0.72143387584947, 0.806704085320234, 0.273651231778786, 0.534617710625753, 0.883442011661828, 0.51143297040835, 0.635279031237587, 0.982488411245868, 0.128191662253812, 0.608250334393233, 0.00211617001332343, 0.805255060316995, 0.323773808777332, 0.306677564512938, 0.937946697231382, 0.952901545912027, 0.356539872474968, 0.924458616180345, 0.578646255191416, 0.17053873417899, 0.677534933434799, 0.324038942577317, 0.823643049690872, 0.299759787274525, 0.00245041307061911, 0.940176057163626, 0.987200211500749, 0.889984066365287, 0.577220825478435, 0.483797369059175, 0.456700396491215, 0.644025508780032, 0.862182233016938, 0.583210124168545, 0.41609926847741, 0.710130138788372, 0.570542151806876, 0.5935685839504, 0.414217859040946, 0.739769909298047, 0.843342585489154, 0.120304950978607, 0.831373705528677, 0.171291669597849, 0.0735178210306913, 0.030190245481208, 0.558681540191174, 0.644223501672968, 0.494808136951178, 0.556333896936849, 0.0986431480851024, 0.531548792263493, 0.669469459913671, 0.520769107853994, 0.404627975076437, 0.285342098912224, 0.434059143532068, 0.49357670894824, 0.361874393420294, 0.886208007112145, 0.608362445840612, 0.779293450992554, 0.791784947738051, 0.0354120472911745, 0.0831993385218084, 0.526240041712299, 0.651936516631395, 0.0503480019979179, 0.823625079123303, 0.279316153842956, 0.402047312352806, 0.219654568005353, 0.618777625029907, 0.0802716456819326, 0.12616874743253, 0.999882121803239, 0.0887981646228582, 0.730766302207485, 0.0184221812523901, 0.667357108322904, 0.411607096670195, 0.840884877135977, 0.0269842050038278, 0.0756081184372306, 0.103879877598956, 0.696790269110352, 0.166358743794262, 0.265813338337466, 0.748592962045223, 0.678929446730763, 0.300610319944099, 0.241655539721251, 0.707292702747509, 0.203237793641165, 0.892184275900945, 0.693464714800939, 0.277054191567004, 0.936139681376517, 0.136029158253223, 0.119701212504879, 0.527949186740443, 0.159549456089735, 0.82778755389154, 0.202289196662605, 0.78968838788569, 0.911913087591529, 0.688663375098258, 0.539220720063895, 0.49137938930653, 0.876462150365114, 0.402406141161919, 0.0575403196271509, 0.778579277452081, 0.628172406228259, 0.690493760863319, 0.33758073952049, 0.899720995686948, 0.0511564207263291, 0.635021428344771, 0.118946098722517, 0.674115343252197, 0.833214801736176, 0.693689559353516, 0.999731228454039, 0.246809442061931, 0.638021025341004, 0.706278619589284, 0.800849076360464, 0.219591536326334, 0.192402994027361, 0.978462682338431, 0.598387941950932, 0.874444140354171, 0.608039286453277, 0.681696287123486, 0.05446284590289, 0.490601705852896, 0.68625046662055, 0.420140779810026, 0.846875556046143, 0.589140177471563, 0.137673994991928, 0.706202771514654, 0.520421259338036, 0.403482888359576, 0.967488228110597, 0.967913971515372, 0.653840645216405, 0.679689550306648, 0.232633358100429, 0.625390260014683, 0.165108924731612, 0.415078071178868, 0.642243389505893, 0.173946996452287, 0.883843315998092, 0.762139811879024, 0.617300033569336, 0.822709401836619, 0.981717861024663, 0.318369371350855, 0.260497168637812, 0.586254798807204, 0.735153288813308, 0.283184831496328, 0.29458187147975, 0.109728863695636, 0.233773864107206, 0.850751985562965, 0.666136714396998, 0.243743690196425, 0.639503082260489, 0.74163541616872, 0.315662161214277, 0.338157590478659, 0.958024427527562, 0.523312629666179, 0.916950100334361, 0.479123265948147, 0.177702227374539, 0.749011717969552, 0.210710386512801, 0.693312834016979, 0.264493470545858, 0.850399568444118, 0.84082739171572, 0.531869689468294, 0.130052509019151, 0.941359552554786, 0.315827787388116, 0.0643860918935388, 0.432687648804858, 0.32548119337298, 0.470499201212078, 0.286712027387694, 0.0589737044647336, 0.257354091852903, 0.709207796957344, 0.471231093397364, 0.76849884330295, 0.373975105350837, 0.0765411495231092, 0.397612841567025, 0.620226120576262, 0.511317259632051, 0.7325252040755, 0.264136401005089, 0.752043495653197, 0.575241702608764, 0.696145281894132, 0.668385970871896, 0.833036744734272, 0.0644992284942418, 0.724380935775116, 0.0302554867230356, 0.375992152839899, 0.256838595494628, 0.963381164474413, 0.572680699173361, 0.700566933490336, 0.674056420801207, 0.0268750940449536, 0.257733263541013, 0.499936993932351, 0.736343321856111, 0.695421858690679, 0.375591075513512, 0.0955182849429548, 0.855028142686933, 0.879141342826188, 0.178740677423775, 0.628038023132831, 0.417502142256126, 0.0284669087268412, 0.872229226864874, 0.386717851972207, 0.0690116935875267, 0.0325202394742519, 0.853756631026044, 0.68649323657155, 0.0661306548863649, 0.620747390436009, 0.980441068299115, 0.438696285476908, 0.658167254878208, 0.326496475376189, 0.183312180219218, 0.825756628531963, 0.0748140914365649, 0.0827272285241634, 0.794643523171544, 0.0681562682148069, 0.770668093347922, 0.631532128201798, 0.9328072310891, 0.061944525456056, 0.167062664637342, 0.108632492832839, 0.917237809859216, 0.0848774816840887, 0.374729287112132, 0.642951080575585, 0.39102722518146, 0.100814888253808, 0.84228875534609, 0.592602321412414, 0.727506822440773, 0.442961059510708, 0.675748377805576, 0.69502053107135, 0.0118908530566841, 0.479887200985104, 0.542015607701614, 0.732864574063569, 0.351494233123958, 0.881945214699954, 0.669266083976254, 0.987502825912088, 0.528625160921365, 0.0667903940193355, 0.00140442862175405, 0.242058286443353, 0.869482255075127, 0.826868113130331, 0.0137656715232879, 0.734295883681625, 0.167530622566119, 0.760752833681181, 0.292518760776147, 0.20499760680832, 0.823405368486419, 0.443984432844445, 0.12052469397895, 0.570051037473604, 0.516148637747392, 0.910295703448355, 0.537475473247468, 0.308267738902941, 0.677415958140045, 0.272365507204086, 0.06837368151173, 0.623551483731717, 0.379867600742728, 0.44568228116259, 0.715087722986937, 0.608884535962716, 0.697632605908439, 0.855405785841867, 0.260610054479912, 0.821308285230771, 0.1006129081361, 0.288011169061065, 0.160594563698396, 0.130569245666265, 0.47789107891731, 0.199056484503672, 0.489000813104212, 0.522503468906507, 0.00345114967785776, 0.895369889913127, 0.156179877929389, 0.896554252132773, 0.0051896283403039, 0.159498208435252, 0.801389569183812, 0.425538387149572, 0.76255826628767, 0.485525158233941, 0.930045259650797, 0.938285271637142, 0.303899776656181, 0.572766352212057, 0.494353072252125, 0.762224508915097, 0.49339178390801, 0.656807065010071, 0.62745642894879, 0.805895890342072, 0.486195153091103, 0.799294675933197, 0.738624058198184, 0.910582388984039, 0.493776180082932, 0.763800260145217, 0.864176835864782, 0.262446246575564, 0.50154186738655, 0.735291791614145, 0.0717381495051086, 0.514736128039658, 0.420928674284369, 0.486797855235636, 0.267150771804154, 0.346612408058718, 0.802851508604363, 0.268189293099567, 0.236960903275758, 0.478872953215614, 0.458300088066608, 0.490496910642833, 0.592086177552119, 0.852017582627013, 0.525441447272897, 0.282234843354672, 0.129057081183419, 0.290590130258352, 0.344687791541219, 0.828081331448629, 0.376178169157356, 0.617703998461366, 0.13803172041662, 0.155742635717615, 0.203340109903365, 0.404067793861032, 0.164989845361561, 0.298154953867197, 0.0996939458418638, 0.486985788447782, 0.816511142300442, 0.992121336748824, 0.165002305759117, 0.925647258525714, 0.749066040851176, 0.548747672466561, 0.508810407947749, 0.801786865107715, 0.853415144607425, 0.393846796592698, 0.274730474455282, 0.0955638345330954, 0.112292376346886, 0.607012189226225, 0.529940035194159, 0.713753649033606, 0.711379387183115, 0.137443387648091, 0.733451207634062, 0.651168831624091, 0.823772669536993, 0.966875478858128, 0.619537439662963, 0.263459998648614, 0.698784355772659, 0.843732595443726, 0.153750819154084, 0.249574399320409, 0.69652198953554, 0.870144786080346, 0.0481305280700326, 0.122865529265255, 0.249022350879386, 0.469056337373331, 0.299708938924596, 0.516280655981973, 0.71453296020627, 0.310235685901716, 0.125962223159149, 0.0660402823705226, 0.695044700521976, 0.352761020418257, 0.754392564529553, 0.48133332352154, 0.48866971465759, 0.180001656524837, 0.241779192583635, 0.0218009320087731, 0.864320070948452, 0.710841834545135, 0.966302966699004, 0.250015857396647, 0.103905845666304, 0.935055992566049, 0.773843439528719, 0.304070895537734, 0.639022790826857, 0.174918729811907, 0.577328010695055, 0.795260609360412, 0.999415898462757, 0.454457530053332, 0.810331338085234, 0.508056019898504, 0.20717213046737, 0.73664964386262, 0.463632893050089, 0.203475865302607, 0.233904311666265, 0.817307952558622, 0.461816230090335, 0.927997923921794, 0.469560488592833, 0.587983972160146, 0.0068761701695621, 0.333074032561854, 0.997285416116938, 0.967299019685015, 0.670155686326325, 0.0866130641661584, 0.895532762631774, 0.168180779088289, 0.953933055046946, 0.780981434276327, 0.111224338877946, 0.0143611053936183, 0.271563645685092, 0.595045728608966, 0.988186666509137, 0.419084780151024, 0.438135511241853, 0.494399385061115, 0.39125221176073, 0.113945916760713, 0.549806158524007, 0.788047116948292, 0.420482554240152, 0.68601179565303, 0.764733287738636, 0.0890646623447537, 0.0858459821902215, 0.792576586361974, 0.213338974397629, 0.485717610921711, 0.918992176651955, 0.736867806641385, 0.664923338685185, 0.536157767754048, 0.247778926510364, 0.0271996930241585, 0.908870942657813, 0.297028697561473, 0.307224867166951, 0.939665579935536, 0.0738629472907633, 0.913117202697322, 0.342939376132563, 0.44183403160423, 0.427788317902014, 0.743232470471412, 0.299446316668764, 0.934799679554999, 0.297009933739901, 0.425794660346583, 0.770876366877928, 0.404343879316002, 0.888593745185062, 0.688565081218258, 0.936833801213652, 0.863029075553641, 0.138467714190483, 0.672391059575602, 0.175889304373413, 0.2948433898855, 0.139533770736307, 0.997636652551591, 0.764447713736445, 0.941279981052503, 0.395430647302419, 0.988649593899027, 0.860932387877256, 0.251324668526649, 0.543680759845302, 0.864005510928109, 0.596790359122679, 0.765046139713377, 0.545096346177161, 0.403101695468649, 0.718592434888706, 0.0331420467700809, 0.726242486387491, 0.957839703187346, 0.380871819565073, 0.239212960004807, 0.626043035183102, 0.554762750631198, 0.494491616263986, 0.178449619095773, 0.237806958844885), actual = c(1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.262386780930683, 0.084515308495611, 0.161723572295159, 0.0900105126202106, 0.901959513546899, 0.806039406917989, 0.957384396810085, 0.0626209487672895, 0.962034896947443, 0.168174240272492, 0.148664253298193, 0.918630411149934, 0.779766591265798, 0.998899067752063, 0.921921357279643, 0.941878670826554, 0.392853266559541, 0.179213860537857, 0.00497308000922203, 0.607120845466852, 0.259821520885453, 0.990187602117658, 0.157316455617547, 0.983816655119881, 0.352246471913531, 0.078906548442319, 0.531557901762426, 0.496252557029948, 0.980456771096215, 0.371282974956557, 0.728935353457928, 0.66453061089851, 0.534690904663876, 0.97496030270122, 0.372794247465208, 0.417940252460539, 0.534505610354245, 0.447868959279731, 0.47854878869839, 0.729706063168123, 0.213357131695375, 0.416571632493287, 0.0808800384402275, 0.980365202063695, 0.0731516003143042, 0.370085487607867, 0.588561230804771, 0.361782822757959, 0.622194684576243, 0.20432276208885, 0.910713557619601, 0.87417414970696, 0.72021149424836, 0.989535521715879, 0.658441791078076, 0.359882720978931, 0.652349872281775, 0.0953674900811166, 0.313493624329567, 0.222656313329935, 0.410084087168798, 0.998037951299921, 0.404329432407394, 0.586077573476359, 0.105774727184325, 0.591041296953335, 0.54121155734174, 0.945572202559561, 0.44962270790711, 0.281309681478888, 0.0257248885463923, 0.923354994272813, 0.778537089470774, 0.799433608073741, 0.450362112838775, 0.455000252230093, 0.0330902501009405, 0.828867162112147, 0.552270599408075, 0.973952373024076, 0.950983985327184, 0.0732248101849109, 0.104501688852906, 0.33149245637469, 0.972912734141573, 0.0870303967967629, 0.44894879614003, 0.885471222689375, 0.774667457211763, 0.741356322541833, 0.198296146700159, 0.799911545589566, 0.34865762735717, 0.233039422659203, 0.486151515971869, 0.997264319797978, 0.486387007171288, 0.121077900519595, 0.835452264873311, 0.294905642746016, 0.995574224507436, 0.197074103634804, 0.127132503315806, 0.251979334512725, 0.652798632858321, 0.0349180090706795, 0.407334310701117, 0.605778784956783, 0.54706974606961, 0.693730738013983, 0.555907142814249, 0.797167909797281, 0.334466222673655, 0.153894388349727, 0.600137347588316, 0.475231430027634, 0.629678996047005, 0.889024471864104, 0.618904160568491, 0.0337464353069663, 0.616258240537718, 0.0942822066135705, 0.314274370204657, 0.655493363039568, 0.17455585882999, 0.752141170669347, 0.506551673868671, 0.572884853696451, 0.797725641867146, 0.429715216159821, 0.055439563235268, 0.786207997705787, 0.856555512174964, 0.364337561652064, 0.316490503493696, 0.97447248804383, 0.872360995970666, 0.077087661717087, 0.349672610173002, 0.0367631623521447, 0.458080301526934, 0.00493083987385035, 0.129825858399272, 0.540652649244294, 0.0993892690166831, 0.682982751401141, 0.184321971843019, 0.775941821280867, 0.267123489407822, 0.848059945972636, 0.613957373891026, 0.402945926878601, 0.221849500201643, 0.78284637350589, 0.19857124472037, 0.845266249962151, 0.669465055922046, 0.797965086996555, 0.854232703568414, 0.524426823714748, 0.641661847243086, 0.796491781249642, 0.463063244707882, 0.785311961779371, 0.660468023968861, 0.689645551145077, 0.284174495376647, 0.380020624026656, 0.60536279133521, 0.835418728645891, 0.722534844186157, 0.848675917601213, 0.405830725561827, 0.657789901597425, 0.0881592556834221, 0.215755976969376, 0.488120166584849, 0.13912149053067, 0.370554760331288, 0.315368404146284, 0.980893091997132, 0.102819693041965, 0.40899472264573, 0.801307286135852, 0.8503816020675, 0.348126879194751, 0.203323554247618, 0.992032035253942, 0.19825791940093, 0.795567768160254, 0.492142525967211, 0.525521634379402, 0.0732827889733016, 0.9573371312581, 0.848096396308392, 0.526161979185417, 0.818350362125784, 0.22616484737955, 0.688756915507838, 0.751661848276854, 0.511988620273769, 0.513065227773041, 0.209353809943423, 0.350602719467133, 0.47593957465142, 0.175823465222493, 0.758445119019598, 0.995222702156752, 0.489728952292353, 0.262237552087754, 0.337200089823455, 0.678885203087702, 0.590350482612848, 0.620659900130704, 0.579337225994095, 0.313459317898378, 0.246639124583453, 0.0194335938431323, 0.469633047701791, 0.170598363270983, 0.884068963583559, 0.752653720322996, 0.493993895594031, 0.588871309999377, 0.8217800243292, 0.787484588101506, 0.448090587044135, 0.860896018566564, 0.983395083341748, 0.934272186597809, 0.488749710144475, 0.221938820090145, 0.0174433109350502, 0.0255233326461166, 0.0923047005198896, 0.318238081876189, 0.48010730324313, 0.331302900332958, 0.140243236208335, 0.978039750130847, 0.554518750403076, 0.464057420380414, 0.576773422537372, 0.62393673392944, 0.665272643789649, 0.267499245004728, 0.794653743738309, 0.0340615420136601, 0.976851076819003, 0.67591408197768, 0.695302707608789, 0.933827622327954, 0.730658700456843, 0.518932981183752, 0.391294485889375, 0.313256659777835, 0.193150090519339, 0.153749259654433, 0.161806893767789, 0.559222050011158, 0.875783078372478, 0.127470220206305, 0.0908722202293575, 0.324514608364552, 0.845870316261426, 0.937605147482827, 0.899088061414659, 0.0396348985377699, 0.561160773271695, 0.877968227025121, 0.985077983234078, 0.873047748347744, 0.1859103331808, 0.173272078391165, 0.686911139171571, 0.997792555950582, 0.563792361645028, 0.271702653262764, 0.103869698010385, 0.955278928391635, 0.174743392970413, 0.956710828701034, 0.149387069744989, 0.959043318871409, 0.090371846454218, 0.612030883319676, 0.222356032812968, 0.316588179441169, 0.622915230458602, 0.0868075184989721, 0.561221733922139, 0.190361491870135, 0.486594904446974, 0.278817866928875, 0.824339857324958, 0.240326150786132, 0.662112916586921, 0.219849209766835, 0.81359449471347, 0.926828833995387, 0.743145492626354, 0.382549520581961, 0.597764094360173, 0.872839750954881, 0.948148960247636, 0.81893941340968, 0.532381260767579, 0.552169520407915, 0.806461536791176, 0.395366491749883, 0.714267077157274, 0.0477585617918521, 0.243130758637562, 0.0753069645725191, 0.515774681931362, 0.985873862402514, 0.191281394101679, 0.926840648520738, 0.601228954736143, 0.900500318966806, 0.471011376474053, 0.909443204756826, 0.303561794338748, 0.427009019535035, 0.183251871261746, 0.040289209689945, 0.833940219134092, 0.371661390177906, 0.70658911135979, 0.138803876470774, 0.99408606835641, 0.237405819119886, 0.399288846645504, 0.23259346652776, 0.0784446683246642, 0.407225968781859, 0.982483498752117, 0.351664267713204, 0.114847755292431, 0.309879676438868, 0.0387385885696858, 0.47569874371402, 0.319049141835421, 0.445639270823449, 0.504840266890824, 0.810987825971097, 0.812488957773894, 0.0255498278420419, 0.555598391918465, 0.156316871056333, 0.942818260984495, 0.505020781420171, 0.262206088751554, 0.641022976720706, 0.238080794923007, 0.558992649661377, 0.887062523979694, 0.70386417163536, 0.444781256839633, 0.246925124665722, 0.996629619738087, 0.936603154754266, 0.899552606744692, 0.623486925149336, 0.914720171829686, 0.75884832511656, 0.805526430951431, 0.72872546967119, 0.940948525210842, 0.512690936215222, 0.0377030987292528, 0.942184658488259, 0.287931678583845, 0.8046947598923, 0.0158801025245339, 0.98243893426843, 0.820892649935558, 0.0448738287668675, 0.257864015409723, 0.778441091999412, 0.0570018030703068, 0.197060398524627, 0.274824602995068, 0.0125279743224382, 0.0472045165952295, 0.887474731076509, 0.779286141041666, 0.748131256783381, 0.0150390635244548, 0.4434849685058, 0.96797219500877, 0.128041680669412, 0.99555249675177, 0.0373760284855962, 0.507368197897449, 0.743150971829891, 0.243887177668512, 0.680602357722819, 0.762336208950728, 0.733493940671906, 0.830377005040646, 0.0208988555241376, 0.71864019962959, 0.118784222751856, 0.136843047337607, 0.986510574351996, 0.424885897431523, 0.279288472840562, 0.0478753540664911, 0.681410069344565, 0.15886911121197, 0.557860078522936, 0.278014043578878, 0.395254920236766, 0.359097301028669, 0.155457576503977, 0.174211772857234, 0.538422560319304, 0.577123707393184, 0.750258956104517, 0.56524174194783, 0.549016345059499, 0.863233744632453, 0.552735717268661, 0.47614214499481, 0.330757756251842, 0.780126126250252, 0.600650922162458, 0.0349996066652238, 0.495378117077053, 0.706789171556011, 0.350617555668578, 0.441552172647789, 0.538594498066232, 0.645038752350956, 0.345977107062936, 0.727954105706885, 0.948374880710617, 0.445424120174721, 0.171124819200486, 0.181234163697809, 0.0740055218338966, 0.571246309904382, 0.209341751411557, 0.626171935349703, 0.978019884554669, 0.165722036967054, 0.153124599950388, 0.054666904034093, 0.147841142956167, 0.706667583668604, 0.0922884952742606, 0.226327968761325, 0.651933243265375, 0.986834570998326, 0.790813960367814, 0.808106104610488, 0.321189792826772, 0.173288638005033, 0.156679427949712, 0.0344768003560603, 0.168852944392711, 0.883211591280997, 0.275500506395474, 0.656634303508326, 0.0158713674172759, 0.790294300066307, 0.104995221598074, 0.447368372930214, 0.470643047941849, 0.368436228949577, 0.72132982336916, 0.749859031522647, 0.355370247736573, 0.470126413274556, 0.949405918596312, 0.716723634395748, 0.0586631603073329, 0.5020200854633, 0.506885423092172, 0.780244760448113, 0.0331218242645264, 0.728937714360654, 0.898926484631374, 0.32867429824546, 0.347458918578923, 0.292447566520423, 0.702835784060881, 0.404147952096537, 0.423858492169529, 0.0573904018383473, 0.252877961145714, 0.425037170527503, 0.22965128836222, 0.472936321282759, 0.00458378181792796, 0.286813481943682, 0.727707374608144, 0.351568278158084, 0.534939419012517, 0.462419619085267, 0.592061291681603, 0.495016921311617, 0.652214674046263, 0.36895401426591, 0.88610239746049, 0.659052410861477, 0.223569157766178, 0.319295727647841, 0.428450189298019, 0.934823353309184, 0.419497936265543, 0.389877091161907, 0.353305216878653, 0.586322033312172, 0.385141553590074, 0.769660849589854, 0.4140580361709, 0.829511378891766, 0.800504203420132, 0.594526099506766, 0.773645834997296, 0.284832368604839, 0.967403171816841, 0.754633887438104, 0.478756390279159, 0.00990165676921606, 0.679469489958137, 0.0714047604706138, 0.83765982161276, 0.935693464474753, 0.197294412646443, 0.500385238556191, 0.905072666937485, 0.910982593894005, 0.452864626888186, 0.871277169557288, 0.639540096744895, 0.704553483752534, 0.866676893318072, 0.696259528398514, 0.837996957823634, 0.957399070030078, 0.110682397149503, 0.0878688611555845, 0.524788608308882, 0.252224464900792, 0.798123746644706, 0.450939664850011, 0.302069288212806, 0.670446182601154, 0.665740203578025, 0.0535857656504959, 0.252241212176159, 0.696169807575643, 0.210697015514597, 0.454855381045491, 0.0930070779286325, 0.0328497104346752, 0.161574819358066, 0.635287873679772, 0.355052541941404, 0.78706103656441, 0.195988588966429, 0.290104393381625, 0.560693588107824, 0.565443343715742, 0.780084165278822, 0.411999496631324, 0.491139994235709, 0.577148980693892, 0.83731228671968, 0.885829238221049, 0.103958390885964, 0.226625525625423, 0.692742934217677, 0.645902749383822, 0.133550752419978, 0.348926992621273, 0.103102955035865, 0.134023846592754, 0.729591205250472, 0.104283469961956, 0.691282881423831, 0.238687727833167, 0.593304899288341, 0.0275785000994802, 0.426528856623918, 0.100815205834806, 0.268175693694502, 0.640493501676247, 0.577833281364292, 0.653548459522426, 0.250322918407619, 0.945744836470112, 0.752588596660644, 0.0802248397376388, 0.582485216902569, 0.604454084765166, 0.554541930789128, 0.565593748120591, 0.249984892550856, 0.112085608066991, 0.270348079502583, 0.824533071136102, 0.505611374275759, 0.00744358683004975, 0.0888860106933862, 0.337701841956005, 0.845102282240987, 0.736999671906233, 0.788382844533771, 0.4595775839407, 0.422688645077869, 0.458255138946697, 0.451112817972898, 0.606039801612496, 0.518975311191753, 0.0641232533380389, 0.788766512414441, 0.957504263380542, 0.964906667592004, 0.444629054050893, 0.346698329318315, 0.156945213209838, 0.585869582369924, 0.0747498120181262, 0.423153237905353, 0.0808112907689065, 0.0275267150718719, 0.839965735794976, 0.996722770389169, 0.343842639354989, 0.591100110439584, 0.569214179646224, 0.595301588531584, 0.360277919331565, 0.00499317073263228, 0.160342851420864, 0.839577511884272, 0.86340120434761, 0.671676211524755, 0.782019364880398, 0.656793644186109, 0.453330823453143, 0.93819379247725, 0.208760421723127, 0.654105969239026, 0.671931218355894, 0.575308786006644, 0.596564129227772, 0.850378400878981, 0.551557332510129, 0.0294072555843741, 0.906835091300309, 0.211420848732814, 0.0830351393669844, 0.398580658482388, 0.898384175263345, 0.265485676703975, 0.338810254586861, 0.0576382055878639, 0.455354580190033, 0.992705856915563, 0.728505665436387, 0.374173861229792, 0.962685456266627, 0.43977761361748, 0.684207923011854, 0.628319063223898, 0.254287941614166, 0.165991526795551, 0.336765531217679, 0.85755282570608, 0.606267990544438, 0.891711992910132, 0.753826897591352, 0.595996075309813, 0.461642561713234, 0.590695437509567, 0.34728867886588, 0.560664959251881, 0.513234854675829, 0.427797601325437, 0.368708779336885, 0.568910088855773, 0.436235769186169, 0.561184180434793, 0.93067888286896, 0.214109007269144, 0.761625936022028, 0.252480695489794, 0.423771331785247, 0.888079478405416, 0.0340268791187555, 0.889617040986195, 0.444726945599541, 0.224557796726003, 0.859427594812587, 0.632516703335568, 0.470837669214234, 0.585986748337746, 0.817682188935578, 0.300617761909962, 0.303188608959317, 0.622765136417001, 0.506713985232636, 0.444426201051101, 0.146385811967775, 0.94785038032569, 0.749790793051943, 0.792044499889016, 0.337796351639554, 0.0181817759294063, 0.553544320166111, 0.584579112008214, 0.850678088609129, 0.99775189999491, 0.0360951328184456, 0.367534815566614, 0.0594091969542205, 0.553502799943089, 0.250676446361467, 0.489203249337152, 0.40094920177944, 0.0319224069826305, 0.327109065139666, 0.710433664731681, 0.738398217130452, 0.935236630728468, 0.150912540499121, 0.43950427393429, 0.624319661641493, 0.855516185984015, 0.999971964629367, 0.680054351454601, 0.718890810385346, 0.15082442201674, 0.734645303571597, 0.566562857711688, 0.209886056603864, 0.183902490651235, 0.726467673201114, 0.294111010618508, 0.619950442109257, 0.375286115100607, 0.66913097538054, 0.687816750956699, 0.630792600335553, 0.516882351366803, 0.00142253469675779, 0.321434057783335, 0.783282197313383, 0.835240971064195, 0.318899057805538, 0.0816940481308848, 0.468634170712903, 0.973678993526846, 0.456987529527396, 0.582930085947737, 0.739285232266411, 0.130309691186994, 0.95951510197483, 0.954586544306949, 0.0534092241432518, 0.697335695382208, 0.677129698684439, 0.735127745661885, 0.937452772399411, 0.638302251929417, 0.192834552144632, 0.0258787022903562, 0.414758285973221, 0.312649975065142, 0.0197176560759544, 0.326075763674453, 0.608107958454639, 0.613898234441876, 0.753040592651814, 0.642922609811649, 0.162517778109759, 0.94542717305012, 0.716185760451481, 0.337423938093707, 0.248292545089498, 0.0812088705133647, 0.6976041670423, 0.257122728740796, 0.989458885509521, 0.509748408570886, 0.880237092729658, 0.500350161921233, 0.597517158603296, 0.543569572502747, 0.0428412342444062, 0.24387287418358, 0.724159812321886, 0.801701860735193, 0.0418911050073802, 0.39910545386374, 0.854462993564084, 0.670573278330266, 0.982605017023161, 0.704168153461069, 0.294915578095242, 0.909190063131973, 0.585711371619254, 0.961935839150101, 0.297173532890156, 0.511396378744394, 0.154463920276612, 0.599370906827971, 0.750122997909784, 0.429349037120119, 0.826651263283566, 0.503643405158073, 0.88783454708755, 0.32157915411517, 0.701666510663927, 0.348852583672851, 0.32258341065608, 0.177242990816012, 0.239781088195741, 0.756788419326767, 0.387086629169062, 0.880493145436049, 0.20297941355966, 0.354447788093239, 0.112610667943954, 0.549566744593903, 0.961187043925747, 0.978688515489921, 0.752201313152909, 0.313184487866238, 0.141652257414535, 0.566147637553513, 0.133662126027048, 0.160747094079852, 0.395722579443827, 0.651642334414646, 0.0772322015836835, 0.760609242599458, 0.395179946441203, 0.304049858124927, 0.697988042607903, 0.358003166504204, 0.366831536171958, 0.998704689787701, 0.713394450722262, 0.967053080210462, 0.956366745755076, 0.88801887864247, 0.255111598176882, 0.601018954301253, 0.806518291588873, 0.5928805898875, 0.211959681706503, 0.552830260712653, 0.424106862861663, 0.710823157569394, 0.687454009894282, 0.956102226860821, 0.473799527389929, 0.332273210864514, 0.785121944965795, 0.3993781642057, 0.818462044931948, 0.292609907221049, 0.0216086031869054, 0.00920310476794839, 0.00797977461479604, 0.998969333479181, 0.392419431358576, 0.654926831601188, 0.513473294209689, 0.847859155386686, 0.930561369750649, 0.17109953565523, 0.885294067906216, 0.0132899803575128, 0.91486354987137, 0.196395028848201, 0.163834881270304, 0.265427716542035, 0.0281370212323964, 0.988702903268859, 0.858140686526895, 0.205177323194221, 0.597213352797553, 0.371090531116351, 0.975534251891077, 0.468526809941977, 0.112860254943371, 0.878844901220873, 0.468844131333753, 0.81548182386905, 0.370295310625806, 0.279146840563044, 0.754684318555519, 0.690962457098067, 0.51870446302928, 0.503233122872189, 0.899916118243709, 0.647229727124795, 0.131396748125553, 0.27001745486632, 0.994652724592015, 0.331078081624582, 0.632932618027553, 0.477768676821142, 0.112959227291867, 0.959561447612941, 0.378249590983614, 0.605491445399821, 0.423496558796614, 0.813436031341553, 0.0364457853138447, 0.854925457853824, 0.480733580887318, 0.880673968233168, 0.844853231217712, 0.620421133004129, 0.636204644339159, 0.714298870647326, 0.735619717976078, 0.513707043137401, 0.0777975283563137, 0.0086767973843962, 0.935890864348039, 0.813188053434715, 0.295580235309899, 0.0244774692691863, 0.573612000793219, 0.367953748907894, 0.970782671589404, 0.699774806154892, 0.236011942848563, 0.89003796945326, 0.659376855939627, 0.606378957163543, 0.116805256577209, 0.880184644833207, 0.647796326084062, 0.134715717751533, 0.399879126111045, 0.457361462991685, 0.126209873473272, 0.73574769240804, 0.438866708893329, 0.505342424148694, 0.802854279289022, 0.72213037405163, 0.130134833510965, 0.954192627454177, 0.740983641706407, 0.0537320841103792, 0.24369878647849, 0.709356521256268, 0.868036675266922, 0.217268782434985, 0.0101630273275077, 0.309581258101389, 0.0237061670050025, 0.230468285270035, 0.453280088724568, 0.629972083261237, 0.68141025537625, 0.0308605185709894, 0.884362615179271, 0.232515539275482, 0.970729167107493, 0.426478089299053, 0.340614321408793, 0.467563134152442, 0.555979973869398, 0.0122276898473501, 0.848387239035219, 0.885377852478996, 0.875563088804483, 0.867080787429586, 0.898614945122972, 0.863717386499047, 0.832111535128206, 0.966805561445653, 0.800120316445827, 0.394530104007572, 0.63870044099167, 0.825606002938002, 0.883573646657169, 0.072694311151281, 0.212740119546652, 0.723542019026354, 0.779184359358624, 0.371603867271915, 0.284507017349824, 0.101403422886506, 0.595411085756496, 0.384106403915212, 0.411901609273627, 0.0394324054941535), actual = c(1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"