Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.6016214        0.6061105 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5085753        0.5085753        0.5164010 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.356055371230468, 0.606350079644471, 0.7241816224996, 0.841250020079315, 0.760536308633164, 0.271621485240757, 0.40918276226148, 0.733244343893602, 0.677428472554311, 0.328470381908119, 0.745852379361168, 0.771986554609612, 0.340593155007809, 0.325986806070432, 0.804437808459625, 0.437616147566587, 0.43768876278773, 0.0519026832189411, 0.309377346653491, 0.63280632533133, 0.221127233933657, 0.310136577347293, 0.13889517262578, 0.535575598478317, 0.478874958353117, 0.0371746914461255, 0.679664786439389, 0.0886015314608812, 0.240567717235535, 0.395323892589658, 0.55409748153761, 0.442561495117843, 0.301761355483904, 0.820924871135503, 0.181881740689278, 0.657079559750855, 0.596578537952155, 0.739024729002267, 0.567784404149279, 0.896666697692126, 0.117730355355889, 0.306729074101895, 0.151853907620534, 0.186020570807159, 0.880247130291536, 0.718714081216604, 0.630608203122392, 0.00872316304594278, 0.853626160183921, 0.284628922585398, 0.718993520131335, 0.00581117114052176, 0.662516307784244, 0.260883098235354, 0.561306667514145, 0.0415500693488866, 0.649563554208726, 0.45708180218935, 0.915654458105564, 0.897813887102529, 0.828399720368907, 0.19466118211858, 0.794430844020098, 0.998800795758143, 0.490448304684833, 0.671453186776489, 0.258877883898094, 0.00524517986923456, 0.816550904419273, 0.425409287214279, 0.699276587227359, 0.0883890800178051, 0.963772167917341, 0.00646614353172481, 0.153464797884226, 0.627778565976769, 0.257043327670544, 0.287395582068712, 0.50750715797767, 0.91107575991191, 0.602260861778632, 0.233287331648171, 0.339707530569285, 0.459111635107547, 0.560232871444896, 0.832337581086904, 0.0300412373617291, 0.744387018028647, 0.392531416611746, 0.289612718392164, 0.0542227500118315, 0.184697166318074, 0.090061076451093, 0.576113318558782, 0.0480058686807752, 0.547251320444047, 0.935139906127006, 0.450270762434229, 0.339924596482888, 0.0589064455125481, 0.0440762247890234, 0.996259639970958, 0.271291345823556, 0.460453239036724, 0.951122898142785, 0.917659336235374, 0.0118360675405711, 0.713218099204823, 0.282997098518535, 0.374017225578427, 0.818151169223711, 0.682790807448328, 0.810890081571415, 0.288033534074202, 0.372958451742306, 0.77393671986647, 0.801852535456419, 0.639346060808748, 0.0904110737610608, 0.556473563425243, 0.372959708794951, 0.492807676782832, 0.902441424084827, 0.969518566969782, 0.904955885140225, 0.733680746285245, 0.152769717853516, 0.0135683219414204, 0.855655638501048, 0.503053674008697, 0.860402051126584, 0.701263835886493, 0.302414709469303, 0.889039130881429, 0.99006652389653, 0.716641117120162, 0.790999549906701, 0.360368012217805, 0.238707603421062, 0.0291081215254962, 0.187891397392377, 0.606784771662205, 0.417977935168892, 0.707767860498279, 0.999737050617114, 0.15810604672879, 0.773154089227319, 0.205731260590255, 0.815080307191238, 0.271406462648883, 0.874435805017129, 0.325714220292866, 0.295902765588835, 0.723887037253007, 0.281143235275522, 0.143371863756329, 0.488332074601203, 0.50819095550105, 0.568185006966814, 0.47464389144443, 0.277277517830953, 0.589361848542467, 0.850888296961784, 0.379949751775712, 0.39864422311075, 0.178711774991825, 0.00760676059871912, 0.88012636336498, 0.708233603741974, 0.207921480992809, 0.995187287218869, 0.867502944543958, 0.952594651374966, 0.501228119013831, 0.360932835377753, 0.179774821037427, 0.86026539793238, 0.067660024156794, 0.293310698121786, 0.549942751647905, 0.298785807099193, 0.518418455729261, 0.51455635484308, 0.524374461034313, 0.322987429331988, 0.171082571381703, 0.303387254709378, 0.170896442839876, 0.941088513005525, 0.0532636693678796, 0.492634080117568, 0.316534944809973, 0.211474461713806, 0.507488064235076, 0.979538744082674, 0.771456387359649, 0.2646756505128, 0.879269919125363, 0.160165606765077, 0.90070671052672, 0.715055091772228, 0.233701287070289, 0.207507699728012, 0.553477133624256, 0.28744742856361, 0.606986079597846, 0.59507576492615, 0.908507424639538, 0.0730275604873896, 0.0717214974574745, 0.164807926863432, 0.655045632272959, 0.484083069488406, 0.629626404028386, 0.811319313710555, 0.84461988764815, 0.867084560450166, 0.528207581024617, 0.962802345165983, 0.726752617862076, 0.290853888727725, 0.500883290311322, 0.253703742753714, 0.578928773291409, 0.696891605388373, 0.169978238642216, 0.188733593560755, 0.271337716840208, 0.691579732578248, 0.585278642363846, 0.66631482867524, 0.99286084366031, 0.221483329311013, 0.31310984888114, 0.397172951605171, 0.152786476304755, 0.362617790000513, 0.671274020569399, 0.79502385458909, 0.146940819686279, 0.404334241757169, 0.68789671594277, 0.105794162722304, 0.252104252343997, 0.801956643816084, 0.903948161751032, 0.93552557262592, 0.669474269030616, 0.833400276722386, 0.606462717289105, 0.565445409156382, 0.448641126044095, 0.689215889200568, 0.795719180488959, 0.495923731010407, 0.376421161694452, 0.838784491410479, 0.753964513307437, 0.235827076714486, 0.482220947276801, 0.392342153936625, 0.0647724303416908, 0.252205435419455, 0.658878644229844, 0.153135586762801, 0.335162100847811, 0.143976890482008, 0.629431526875123, 0.880933515029028, 0.0101251758169383, 0.334961249027401, 0.398053737124428, 0.227264674613252, 0.307088425848633, 0.220199925592169, 0.197105827508494, 0.477515990845859, 0.520353313302621, 0.0893325679935515, 0.535387671785429, 0.914458388928324, 0.674987739417702, 0.982000023825094, 0.633282010210678, 0.862217022338882, 0.353249535663053, 0.00122708268463612, 0.262404774548486, 0.238811279181391, 0.450983155984432, 0.545291256159544, 0.461444038664922, 0.74884898471646, 0.825185284949839, 0.417830941500142, 0.628507298417389, 0.41173456935212, 0.916559573495761, 0.63814058364369, 0.736078025074676, 0.913622251246125, 0.999760134611279, 0.755946586839855, 0.070402019424364, 0.0814474241342396, 0.0948610540945083, 0.915353845339268, 0.900190143380314, 0.826017260784283, 0.957319943234324, 0.994614560855553, 0.280150438658893, 0.568316240562126, 0.999592415755615, 0.496092055458575, 0.914152587531134, 0.294605153147131, 0.0494352434761822, 0.0881860833615065, 0.376575579633936, 0.504621682455763, 0.362871858524159, 0.121709024533629, 0.7148781940341, 0.195454831235111, 0.291637115646154, 0.558333139400929, 0.328853910090402, 0.172196431551129, 0.38812965201214, 0.164496909826994, 0.839513519546017, 0.793294122675434, 0.0688926430884749, 0.29203378316015, 0.228221483062953, 0.252185007324442, 0.704557938268408, 0.618541093776003, 0.0134737524203956, 0.255367909325287, 0.275888098636642, 0.481762626208365, 0.0509186377748847, 0.264943992486224, 0.669597125612199, 0.170884244842455, 0.402842998038977, 0.571554210502654, 0.760160871082917, 0.557190127205104, 0.042376612778753, 0.640636075288057, 0.20954940514639, 0.780649932101369, 0.887120904400945, 0.687515077413991, 0.0369552809279412, 0.364254663232714, 0.858493115520105, 0.41889417078346, 0.636380509240553, 0.197422164259478, 0.829173150705174, 0.309798229951411, 0.594373550731689, 0.721254282165319, 0.899528635898605, 0.0418623322620988, 0.325606352649629, 0.8550449332688, 0.745303301140666, 0.0513555682264268, 0.455693282885477, 0.692837089998648, 0.574846013681963, 0.947863183217123, 0.792157341726124, 0.102984235389158, 0.0310117290355265, 0.501613148488104, 0.499051615130156, 0.0366340428590775, 0.340096650179476, 0.893239228986204, 0.251637275330722, 0.226493052206933, 0.225227238843217, 0.483900242717937, 0.546375739155337, 0.67147907987237, 0.690632005920634, 0.675006669946015, 0.664506169268861, 0.835532438242808, 0.34309467067942, 0.955528127029538, 0.70155917503871, 0.655092068947852, 0.669419545447454, 0.652169631794095, 0.617093666223809, 0.50723334774375, 0.791052398039028, 0.69958472950384, 0.0609058132395148, 0.947929470799863, 0.525222520111129, 0.692029818426818, 0.637614939827472, 0.549669316969812, 0.512176817050204, 0.67736917338334, 0.00140374852344394, 0.168427566532046, 0.909656099742278, 0.504065032349899, 0.374901096336544, 0.132451826706529, 0.46355108032003, 0.725460563786328, 0.975700197741389, 0.670179068343714, 0.937637019669637, 0.498523284681141, 0.581285058055073, 0.921543563716114, 0.355568268103525, 0.413834127830341, 0.862624445697293, 0.362829747609794, 0.368747015949339, 0.593188594095409, 0.819927407195792, 0.202566648134962, 0.916274446528405, 0.619253116426989, 0.941602510632947, 0.118701746687293, 0.922525028930977, 0.266320034861565, 0.653738986002281, 0.337517397711053, 0.612188751343638, 0.0531291363295168, 0.24909809557721, 0.938623693073168, 0.678335261298344, 0.641069499310106, 0.87911059637554, 0.987212258623913, 0.648240111535415, 0.191199749009684, 0.646408921573311, 0.192866429686546, 0.942174374358729, 0.567573289154097, 0.676559976302087, 0.0917943494860083, 0.458659183001146, 0.548993202392012, 0.965075387153774, 0.0103951557539403, 0.0779956036712974, 0.905352983390912, 0.348217899212614, 0.425537960138172, 0.518693577498198, 0.769406638573855, 0.855351918376982, 0.263294916599989, 0.645497441990301, 0.0773602561093867, 0.697721478529274, 0.705057149985805, 0.236434337683022, 0.202150071505457, 0.00914162630215287, 0.20034313807264, 0.0916334874927998, 0.914403870003298, 0.239275416592136, 0.588548379018903, 0.155092735076323, 0.183980533154681, 0.567304712254554, 0.362282795133069, 0.914129628567025, 0.694459593156353, 0.566355755552649, 0.764898879220709, 0.989589470438659, 0.505773225566372, 0.404371013399214, 0.1760460564401, 0.338702565059066, 0.0634836677927524, 0.526131445076317, 0.569795060902834, 0.843700022436678, 0.734761873492971, 0.483923398191109, 0.215011782944202, 0.57731602829881, 0.0488916782196611, 0.304523556493223, 0.210567470872775, 0.438232500338927, 0.43061419413425, 0.534227396827191, 0.889352336293086, 0.151400001952425, 0.320634747622535, 0.9430661580991, 0.446683656191453, 0.58304281393066, 0.32865474675782, 0.381359453080222, 0.521059709368274, 0.0901337000541389, 0.801877241814509, 0.0876625669188797, 0.661157177994028, 0.403161377646029, 0.431443513836712, 0.901218448532745, 0.258379626553506, 0.407155248103663, 0.826407762710005, 0.863607537234202, 0.610623072367162, 0.241429730784148, 0.631926593137905, 0.577835391974077, 0.635895342333242, 0.239258813206106, 0.735422852216288, 0.0927310055121779, 0.110242237802595, 0.794607115443796, 0.407579946098849, 0.789105400210246, 0.644793293206021, 0.439480536384508, 0.629042821936309, 0.010027171112597, 0.224242524011061, 0.0710490115452558, 0.424994840752333, 0.20990795828402, 0.460730282822624, 0.649978291243315, 0.891335228225216, 0.512201622361317, 0.234117667423561, 0.414330935338512, 0.694174222415313, 0.636101979762316, 0.0988552884664387, 0.717842532787472, 0.598110970808193, 0.187058972893283, 0.727152685169131, 0.662166809663177, 0.0111707916948944, 0.61340423184447, 0.487232072977349, 0.144949500914663, 0.502811280079186, 0.942144132917747, 0.451196777634323, 0.527928594034165, 0.597859928384423, 0.611636328976601, 0.411750330822542, 0.664807048393413, 0.19917458225973, 0.93288957234472, 0.211489173118025, 0.0931387529708445, 0.299435645109043, 0.989273906685412, 0.564671646337956, 0.192299340385944, 0.964440160663798, 0.876849878579378, 0.0609947117045522, 0.263201565714553, 0.288627086440101, 0.121916387230158, 0.66546907229349, 0.582466299645603, 0.756310163531452, 0.572546068113297, 0.887859809678048, 0.400068041868508, 0.34881782415323, 0.0321995764970779, 0.725325050065294, 0.550101511878893, 0.469153280835599, 0.601154251722619, 0.907587226945907, 0.918193223886192, 0.00399674545042217, 0.45364680676721, 0.249584600329399, 0.0708297432865947, 0.703394518466666, 0.534566634567454, 0.405101499753073, 0.694481084821746, 0.410790856694803, 0.609124833252281, 0.606050446862355, 0.605906317476183, 0.276912087574601, 0.410787822911516, 0.240656983107328, 0.0796974860131741, 0.625655127223581, 0.662401867564768, 0.784092060988769, 0.886032957583666, 0.545340172480792, 0.263971120351925, 0.514024075120687, 0.357457413338125, 0.495438524056226, 0.876445474801585, 0.70302318315953, 0.711747429333627, 0.355904302094132, 0.232013219036162, 0.23265896528028, 0.779019919922575, 0.942705042660236, 0.788476010086015, 0.15951391775161, 0.614595760591328, 0.209962379653007, 0.300036204978824, 0.655606677290052, 0.652279894100502, 0.886331163812429, 0.56431557959877, 0.358484186697751, 0.509155479026958, 0.610230094986036, 0.0699751924257725, 0.658564416924492, 0.535365478368476, 0.899282455444336, 0.334046577103436, 0.694883055053651, 0.667309842305258, 0.60450313007459, 0.775233512744308, 0.14357551257126, 0.611843085382134, 0.166875022929162, 0.942655988503247, 0.979896218050271, 0.318155422573909, 0.551469704601914, 0.791792143136263, 0.668167170369998, 0.989201944321394, 0.990763164358214, 0.428277274593711, 0.0371043938212097, 0.232122610788792, 0.457039317116141, 0.807756609050557, 0.293782672611997, 0.402144517283887, 0.0517997955903411, 0.239852984901518, 0.726676972582936, 0.604916228214279, 0.371080101002008, 0.30562719819136, 0.467545743798837, 0.554864105768502, 0.317100372398272, 0.0955853248015046, 0.0512794018723071, 0.000262932386249304, 0.156732654199004, 0.217018176103011, 0.218170625157654, 0.556314629968256, 0.56757972529158, 0.73006897023879, 0.105971995508298, 0.566573460586369, 0.767517345258966, 0.627620406914502, 0.681660024216399, 0.0276833926327527, 0.296908693620935, 0.344484607689083, 0.763547168811783, 0.651043637888506, 0.0447208897676319, 0.68704964988865, 0.532112207729369, 0.820236999541521, 0.633025113493204, 0.208942137891427, 0.531872408464551, 0.481771951308474, 0.499470313778147, 0.756100351223722, 0.30795514723286, 0.462810250930488, 0.0969115181360394, 0.588284957921132, 0.683712546946481, 0.756632186472416, 0.962368526263162, 0.885523650562391, 0.95343549689278, 0.0808733582962304, 0.759445275645703, 0.0902016023173928, 0.56972423219122, 0.391186171909794, 0.162680780282244, 0.827256656019017, 0.671618812484667, 0.440943663939834, 0.348823117092252, 0.52212008764036, 0.795026036677882, 0.649747354676947, 0.519036948448047, 0.225659059826285, 0.494087675819173, 0.184299131156877, 0.618043925845996, 0.0466939616017044, 0.853008502395824, 0.21198377548717, 0.486509937560186, 0.963479148922488, 0.225999984657392, 0.201842418638989, 0.00529995909892023, 0.552676223684102, 0.697571193333715, 0.0391321962233633, 0.532234872924164, 0.852970143314451, 0.511255586519837, 0.0843128305859864, 0.461026224074885, 0.952752864686772, 0.206332398112863, 0.156027801102027, 0.532017477788031, 0.341339975595474, 0.485014461213723, 0.269005917012691, 0.195196252316236, 0.0487585295923054, 0.340802695369348, 0.831875797128305, 0.669316753745079, 0.950049784500152, 0.861487171379849, 0.805713825859129, 0.38444247469306, 0.261488192714751, 0.713256089715287, 0.983285120688379, 0.829597481060773, 0.38420938141644, 0.221897831186652, 0.450217665405944, 0.633081672247499, 0.941612001508474, 0.660521753830835, 0.366838960209861, 0.736598542192951, 0.910804198123515, 0.862152144312859, 0.893823718186468, 0.0159672368317842, 0.320008042501286, 0.83835098030977, 0.483856525970623, 0.449063857551664, 0.0871568240690976, 0.560631416272372, 0.304203707026318, 0.272471432806924, 0.0731509972829372, 0.943899491336197, 0.488736856961623, 0.601061112480238, 0.587468602694571, 0.15467567066662, 0.590852146735415, 0.808641018811613, 0.879445104626939, 0.771917515201494, 0.615222331834957, 0.962138702860102, 0.758161929436028, 0.246528128162026, 0.787285017548129, 0.520708479452878, 0.858241413254291, 0.14894900820218, 0.974192142486572, 0.594780884450302, 0.505052671069279, 0.244737244443968, 0.0646625775843859, 0.904069554060698, 0.142680265009403, 0.774347168859094, 0.657977134222165, 0.087752734310925, 0.2138719225768, 0.944198545766994, 0.331967614823952, 0.124336500419304, 0.710708365542814, 0.132402018178254, 0.395971512189135, 0.444084871793166, 0.57590472442098, 0.217865727841854, 0.672363884979859, 0.573576514609158, 0.153261463856325, 0.388119318522513, 0.980307213030756, 0.257057165494189, 0.880305106751621, 0.70339788380079, 0.421220833668485, 0.634796767961234, 0.781262097647414, 0.0986852024216205, 0.190679250983521, 0.130604855716228, 0.649199551902711, 0.288699496071786, 0.406861073570326, 0.931056742323563, 0.700051839696243, 0.442456359742209, 0.872067546937615, 0.845198962371796, 0.15705162123777, 0.832675988320261, 0.122659899294376, 0.692730606533587, 0.653079291339964, 0.395400581648573, 0.992527497000992, 0.508969254326075, 0.368392970878631, 0.20557141629979, 0.878179743187502, 0.137476526666433, 0.599060870474204, 0.301047535380349, 0.514403948560357, 0.927454201504588, 0.717167455470189, 0.430917347082868, 0.00666277599520981, 0.399030515458435, 0.572417089482769, 0.254460622789338, 0.344873788300902, 0.947135120164603, 0.320185749791563, 0.814935825299472, 0.520724071655422, 0.708461077651009, 0.655126356519759, 0.00486991065554321, 0.0294005160685629, 0.201599781168625, 0.760809206636623, 0.543375914450735, 0.93539138766937, 0.924679691670462, 0.141524244099855, 0.574372336734086, 0.347295643528923, 0.194860202725977, 0.310386982979253, 0.994478662032634, 0.90344997914508, 0.000393360387533903, 0.247232404304668, 0.482456675963476, 0.995644884882495, 0.586625864962116, 0.247899185866117, 0.653623018646613, 0.31617468665354, 0.361885640537366, 0.599828298669308, 0.124520398909226, 0.903751491568983, 0.121625510277227, 0.297090467764065, 0.191240553976968, 0.881210154155269, 0.618266027187929, 0.887131936848164, 0.994607196189463, 0.635097567457706, 0.402378103230149, 0.784963675774634, 0.29406037949957, 0.171509605832398, 0.859505275497213, 0.347246125573292, 0.300555334193632, 0.647570261964574, 0.11122723040171, 0.316832915181294, 0.653606455773115, 0.809828767552972, 0.261201929301023, 0.845060667488724, 0.623789683450013, 0.241093515185639, 0.239400786813349, 0.275085197528824, 0.402801122516394, 0.411272761179134, 0.713287342339754, 0.0141464292537421, 0.512658056570217, 0.626137155806646, 0.298027480021119, 0.620357233565301, 0.981418630573899, 0.686004500137642, 0.997318005654961, 0.536557321902364, 0.173202423844486, 0.615493999328464, 0.837984147015959, 0.803258002735674, 0.723773403791711, 0.336713342927396, 0.822363190120086, 0.469315388938412, 0.493185250787064, 0.244246304035187, 0.416123890783638, 0.827510435134172, 0.176151487277821, 0.259348063264042, 0.0899213391821831, 0.175575610715896, 0.391602611169219, 0.556708270218223, 0.154818927869201, 0.830533599946648, 0.211610696977004, 0.920933987479657, 0.451951655792072, 0.140201187459752, 0.212294846307486, 0.944591733627021, 0.438893304904923, 0.426256316481158, 0.152207336621359, 0.352993655949831, 0.553338906494901, 0.214193289866671, 0.744316613767296, 0.171599665656686, 0.588091342942789, 0.373821471584961, 0.880398987792432, 0.850232320604846, 0.569821110460907, 0.461362974718213, 0.628842112608254, 0.177831426728517, 0.945648974739015, 0.305825571063906, 0.902246313402429, 0.871587889268994, 0.819577038520947, 0.791934815468267, 0.150399844860658, 0.617489806609228), actual = c(1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.00556903262622654, 0.405694459099323, 0.388225422240794, 0.130275341682136, 0.79386478359811, 0.952334752306342, 0.372512995265424, 0.26521422015503, 0.846988510573283, 0.397030353546143, 0.701667938614264, 0.671238662907854, 0.0567103340290487, 0.202382837189361, 0.9392367110122, 0.672691721003503, 0.674345348961651, 0.473232289077714, 0.151140871690586, 0.170639139600098, 0.145308884093538, 0.750969066517428, 0.634084755321965, 0.0462617236189544, 0.174248650670052, 0.450506366090849, 0.963967258343473, 0.418354524765164, 0.255547847598791, 0.105196482734755, 0.347943403758109, 0.101805003127083, 0.513653257396072, 0.84804597706534, 0.0850538262166083, 0.0371593844611198, 0.916761419735849, 0.642922741593793, 0.646716533927247, 0.638474072562531, 0.755065714009106, 0.0427773513365537, 0.90906227636151, 0.696793287759647, 0.647504914551973, 0.679238190641627, 0.908849803032354, 0.32950451830402, 0.901148952543736, 0.175547260092571, 0.530165228527039, 0.618008667603135, 0.772266013314947, 0.703884616494179, 0.208970442181453, 0.425465744221583, 0.907097058137879, 0.125492352293804, 0.93478915723972, 0.353787540690973, 0.668752685654908, 0.954329558648169, 0.572502618655562, 0.836427511647344, 0.579286233754829, 0.0177862863056362, 0.254703269805759, 0.122255183989182, 0.92364858253859, 0.589026730740443, 0.439456175081432, 0.407142366748303, 0.0693670914042741, 0.618531276937574, 0.428130148677155, 0.140491045545787, 0.53007991053164, 0.811018420150504, 0.868262650910765, 0.584558088798076, 0.215065475320444, 0.0724995539058, 0.952624658355489, 0.4443615374621, 0.351438992423937, 0.932367357891053, 0.740742492256686, 0.609745079651475, 0.952612434281036, 0.824040544917807, 0.960496521322057, 0.741285746451467, 0.695922468323261, 0.314380050171167, 0.883983512409031, 0.168799726525322, 0.0756253497675061, 0.015429669758305, 0.210796915693209, 0.905294248368591, 0.790808001533151, 0.109933796338737, 0.730305933626369, 0.476159829413518, 0.896587039809674, 0.477792667923495, 0.344342388212681, 0.649908716324717, 0.17195068532601, 0.224664077162743, 0.481501621659845, 0.904866788070649, 0.460325629683211, 0.411222845315933, 0.0263794562779367, 0.485103293787688, 0.555528505239636, 0.229326896136627, 0.673526096623391, 0.125954047543928, 0.0276370940264314, 0.738829317502677, 0.757273548748344, 0.0303237454500049, 0.185191112803295, 0.772319772047922, 0.864115119213238, 0.258278358727694, 0.902235502842814, 0.866571696707979, 0.496746455086395, 0.271319329272956, 0.272344280034304, 0.425847961334512, 0.93004559725523, 0.594940089154989, 0.562882693484426, 0.0759965174365789, 0.389620354631916, 0.547645320417359, 0.32658405858092, 0.324562865542248, 0.834436996141449, 0.97773354803212, 0.63404694525525, 0.765871448209509, 0.279913384001702, 0.57343344273977, 0.298429297748953, 0.543918201467022, 0.778923848411068, 0.487070018192753, 0.449789004633203, 0.923242755001411, 0.0638180372770876, 0.293048164574429, 0.842944415286183, 0.893359975423664, 0.63013495481573, 0.896913789678365, 0.929988137912005, 0.684247826458886, 0.102887927321717, 0.293913342291489, 0.673928120406345, 0.729669756256044, 0.220347950235009, 0.120795486029238, 0.896611233474687, 0.462164090713486, 0.0648743971250951, 0.228531187865883, 0.452120440779254, 0.917433319613338, 0.619923430494964, 0.0654451304581016, 0.667541436385363, 0.549331117654219, 0.0346943479962647, 0.9980374048464, 0.908365038456395, 0.294246301986277, 0.0486208067741245, 0.144806952914223, 0.0707143789622933, 0.916823244420812, 0.860109619097784, 0.193297427380458, 0.35979165439494, 0.628260321682319, 0.715651551727206, 0.316813414683565, 0.539744496112689, 0.361698873341084, 0.996979914139956, 0.404245999176055, 0.0438076481223106, 0.426969098625705, 0.596879790071398, 0.281225028680637, 0.30104136490263, 0.219694190658629, 0.0203075238969177, 0.140498450957239, 0.195401738630608, 0.662834646878764, 0.528739207424223, 0.446566705824807, 0.509807185037062, 0.853565239347517, 0.482626939658076, 0.249061518581584, 0.20834099012427, 0.698440249776468, 0.173766301712021, 0.216035343473777, 0.526947303442284, 0.0244586190674454, 0.389360653003678, 0.442537821130827, 0.341879599029198, 0.280650830594823, 0.229735042667016, 0.679166293004528, 0.219965332886204, 0.801695971051231, 0.615997165208682, 0.147267275024205, 0.409624490188435, 0.811346872942522, 0.338058517780155, 0.677043190924451, 0.470502689480782, 0.17466735211201, 0.737174971960485, 0.440679158084095, 0.0497627723962069, 0.171391861047596, 0.924845891539007, 0.268873513210565, 0.786510740173981, 0.388716756831855, 0.143617020919919, 0.932494381209835, 0.325725920498371, 0.0503447453957051, 0.941066010389477, 0.209511843509972, 0.442340763052925, 0.490245578810573, 0.707914824364707, 0.643544264603406, 0.178719860967249, 0.639532759319991, 0.151180669432506, 0.825892963912338, 0.35899159219116, 0.119120738469064, 0.261550241149962, 0.667935726465657, 0.0251358430832624, 0.45636106771417, 0.329391916515306, 0.758275937987491, 0.471727710915729, 0.119483399670571, 0.541302560362965, 0.234652393730357, 0.470487850252539, 0.464703698642552, 0.427715040044859, 0.405379171250388, 0.776059335330501, 0.708032814087346, 0.317393618403003, 0.862872129539028, 0.1972870323807, 0.85725985863246, 0.449371931608766, 0.0917640284169465, 0.585464542731643, 0.464225188829005, 0.46422426099889, 0.65660706628114, 0.575366736855358, 0.07992516993545, 0.32562822336331, 0.476324415765703, 0.214419052004814, 0.858102998696268, 0.167872985824943, 0.323990112403408, 0.866477199131623, 0.425449785077944, 0.0190751170739532, 0.238923688419163, 0.914246749132872, 0.194528606953099, 0.887618200154975, 0.562652323860675, 0.577400239882991, 0.866626418195665, 0.435241227503866, 0.695128405233845, 0.663083602441475, 0.630004175473005, 0.465249324450269, 0.873061443911865, 0.387754570459947, 0.955044990871102, 0.534023329382762, 0.411919989157468, 0.0864134589210153, 0.650920297484845, 0.787202120758593, 0.545820801053196, 0.844651377061382, 0.728106572292745, 0.250469347229227, 0.788687696214765, 0.854961206670851, 0.918669124599546, 0.37929234537296, 0.983392668189481, 0.154857476707548, 0.471923961304128, 0.859593255678192, 0.659139811992645, 0.970030035823584, 0.373775843763724, 0.33210341911763, 0.516122160246596, 0.605384982656687, 0.705621924716979, 0.341521324589849, 0.644723643781617, 0.740605430677533, 0.0488524616230279, 0.749232881003991, 0.130074318731204, 0.436793839558959, 0.442862137453631, 0.709474057890475, 0.747037353692576, 0.0344620104879141, 0.457108858972788, 0.906488368986174, 0.516432202653959, 0.454848936293274, 0.28716342151165, 0.560924842720851, 0.62548911338672, 0.978456457145512, 0.837508860975504, 0.739326792536303, 0.449903375469148, 0.875212365062907, 0.867230337811634, 0.769457877613604, 0.572380509926006, 0.769974866649136, 0.267982762539759, 0.966784902149811, 0.410751580493525, 0.241957851918414, 0.9081612653099, 0.0142647305037826, 0.42445145919919, 0.178903963649645, 0.753756735473871, 0.471074987668544, 0.843091217335314, 0.573662457754835, 0.923729957547039, 0.782757162582129, 0.431499174563214, 0.579151308862492, 0.0536669092252851, 0.645865872502327, 0.724077580031008, 0.921569936908782, 0.0969621927943081, 0.801370300818235, 0.19008838897571, 0.279664934147149, 0.09288848657161, 0.903278977610171, 0.686156860785559, 0.124401892302558, 0.10284106223844, 0.68192811193876, 0.166868859203532, 0.795648160157725, 0.804582789540291, 0.61821216554381, 0.876610410166904, 0.126563681988046, 0.197008035378531, 0.761086504673585, 0.512686083326116, 0.468076687538996, 0.512334283208475, 0.0924654779955745, 0.156334607861936, 0.710821316810325, 0.325285837287083, 0.591638968093321, 0.873670779867098, 0.67867561429739, 0.329240443650633, 0.515485814539716, 0.565539957256988, 0.300787432352081, 0.873188852565363, 0.537330291001126, 0.29413648811169, 0.130114925093949, 0.0268624436575919, 0.448803234379739, 0.521197903901339, 0.0301436786539853, 0.180384666426107, 0.770494598196819, 0.448795225471258, 0.503090932033956, 0.891562853008509, 0.327437430387363, 0.190418994985521, 0.868005797732621, 0.532166202319786, 0.0841561804991215, 0.465179943712428, 0.563075248617679, 0.123732528183609, 0.426064478000626, 0.0147331533953547, 0.629586431430653, 0.472124705323949, 0.88750554388389, 0.254046266200021, 0.424565200693905, 0.636641093529761, 0.143946667900309, 0.62074432359077, 0.936659366590902, 0.0202538080047816, 0.510954684345052, 0.409426083788276, 0.302311810664833, 0.499335569795221, 0.544603148940951, 0.294790015323088, 0.866027412936091, 0.899545730091631, 0.610887314192951, 0.746122865006328, 0.594108317047358, 0.648579712258652, 0.196158654522151, 0.565222995122895, 0.807313882745802, 0.801652687368914, 0.742620119126514, 0.476783053716645, 0.39315390214324, 0.787648254306987, 0.153055402217433, 0.629557984415442, 0.756846459116787, 0.608541519613937, 0.102450821083039, 0.472291309386492, 0.533091969788074, 0.958227110560983, 0.280080349417403, 0.133813639637083, 0.514294252498075, 0.785018001915887, 0.677594813751057, 0.382239258149639, 0.853906947188079, 0.661482392344624, 0.545419833157212, 0.158667560201138, 0.933040706440806, 0.812561564845964, 0.166806714609265, 0.773940793937072, 0.761314792791381, 0.901865091640502, 0.391954329097643, 0.507550102425739, 0.609674860956147, 0.213812829460949, 0.836436789250001, 0.821350001264364, 0.259000655030832, 0.0828262073919177, 0.104955327231437, 0.923948192503303, 0.172883688705042, 0.168693190207705, 0.137062948429957, 0.0439914613962173, 0.0792138397227973, 0.467317748814821, 0.350397853879258, 0.427360314177349, 0.0487541498150676, 0.44666276848875, 0.382675521541387, 0.369493845384568, 0.134698018897325, 0.969764127163216, 0.250489946687594, 0.0478901248425245, 0.038738087983802, 0.962148478487507, 0.513524271780625, 0.77494822605513, 0.736614368855953, 0.706289170542732, 0.836878899252042, 0.884278449462727, 0.882352937245741, 0.559155392693356, 0.986971734790131, 0.546304409159347, 0.455510577885434, 0.954295795178041, 0.389679282903671, 0.531984403263777, 0.572916898177937, 0.477995801251382, 0.0475566361565143, 0.366302478360012, 0.530721433693543, 0.746368547668681, 0.759366114391014, 0.418361247749999, 0.982216516975313, 0.415655465098098, 0.83552857581526, 0.147876422153786, 0.0470321872271597, 0.600546875270084, 0.634745204122737, 0.960625468520448, 0.6119110067375, 0.300369310891256, 0.273669778369367, 0.718463823199272, 0.956616337411106, 0.967856892384589, 0.0556513061746955, 0.28847777005285, 0.922886471496895, 0.90245327539742, 0.171024292008951, 0.752072575036436, 0.9649341579061, 0.890719659393653, 0.269730712287128, 0.10132555803284, 0.3648537308909, 0.498884121654555, 0.471685119904578, 0.720495489193127, 0.653585813241079, 0.222443297039717, 0.771301995264366, 0.34731480316259, 0.825279279612005, 0.549298979574814, 0.956730807432905, 0.315412098076195, 0.508858203189448, 0.118654159363359, 0.969470899086446, 0.226727245375514, 0.57971613924019, 0.851382507011294, 0.545002053724602, 0.717474789824337, 0.283807364525273, 0.0915789811406285, 0.882046669721603, 0.355092773679644, 0.0623150675091892, 0.102418861351907, 0.622070476412773, 0.0980515033006668, 0.676332824397832, 0.44417417421937, 0.761779933236539, 0.598393171327189, 0.698261784855276, 0.198847214924172, 0.540684826206416, 0.0963194842915982, 0.928865033900365, 0.187015602132306, 0.692184656625614, 0.420191691024229, 0.189557632431388, 0.549508754629642, 0.983900643186644, 0.680475815432146, 0.222961775027215, 0.857523298589513, 0.733355856966227, 0.964561806991696, 0.927418847335503, 0.593319073086604, 0.477891536196694, 0.370662826811895, 0.540683124912903, 0.557752615539357, 0.0517992984969169, 0.431720395805314, 0.50505494675599, 0.835380132310092, 0.304949848912656, 0.27363462257199, 0.344388663303107, 0.462364154402167, 0.706173224840313, 0.816186794778332, 0.707090525189415, 0.305986726190895, 0.699294472113252, 0.178115866845474, 0.859931807266548, 0.791440651752055, 0.620834189467132, 0.0822937975171953, 0.971211688825861, 0.923823634162545, 0.400995372794569, 0.122432318050414, 0.299733547028154, 0.269513135543093, 0.700260700657964, 0.16727919690311, 0.918267957866192, 0.83039289014414, 0.99205305846408, 0.613700991962105, 0.31478119106032, 0.204830585978925, 0.690709101501852, 0.183607433456928, 0.356059241108596, 0.934754483634606, 0.632958900881931, 0.473244129214436, 0.0407133793924004, 0.670715371379629, 0.217912132153288, 0.165951549308375, 0.0838866026606411, 0.0328901608008891, 0.766841262346134, 0.628258785465732, 0.973335811169818, 0.460838053142652, 0.253039485774934, 0.0753903952427208, 0.520834867144004, 0.430724776815623, 0.0717676286585629, 0.258370531490073, 0.1539840050973, 0.760560332797468, 0.423651891294867, 0.296082184184343, 0.066372683737427, 0.169416492572054, 0.562264377018437, 0.940882373368368, 0.684992082882673, 0.407684759935364, 0.327920072013512, 0.302456500241533, 0.876325542107224, 0.352193939732388, 0.385139259742573, 0.713947971584275, 0.0775957864243537, 0.349546652287245, 0.913791717495769, 0.362846436910331, 0.64865132025443, 0.78885054285638, 0.147094717482105, 0.029519273666665, 0.353436350822449, 0.390195307787508, 0.487483302364126, 0.967284497339278, 0.904583968222141, 0.878841404570267, 0.890218668151647, 0.652639013249427, 0.0836247478146106, 0.177218178287148, 0.883849574252963, 0.791614296846092, 0.487220366252586, 0.746098899049684, 0.974193851929158, 0.383697761921212, 0.879243337083608, 0.657190744532272, 0.672827356029302, 0.584072613390163, 0.178952395450324, 0.143303492339328, 0.995399513747543, 0.943120778771117, 0.432035193312913, 0.70582692627795, 0.467869812622666, 0.62289583357051, 0.613728904165328, 0.406694093486294, 0.748288360890001, 0.769284596666694, 0.302522089099512, 0.287444950314239, 0.0575055661611259, 0.482436816673726, 0.855372210498899, 0.218528350815177, 0.700730768498033, 0.244285518769175, 0.589426917955279, 0.913762589683756, 0.871301290113479, 0.170885169878602, 0.87629690929316, 0.235128305619583, 0.161759189562872, 0.914470697985962, 0.43920221622102, 0.271826301002875, 0.720548955956474, 0.867131768027321, 0.0461276094429195, 0.586177566321567, 0.0704005784355104, 0.100260133156553, 0.544227312784642, 0.974993965355679, 0.253944319672883, 0.688097844365984, 0.11954871728085, 0.305703256744891, 0.213486413005739, 0.723644791403785, 0.977784221060574, 0.938453758601099, 0.413896539248526, 0.866112242918462, 0.878955429187045, 0.703067234251648, 0.714065940584987, 0.906935568898916, 0.9922489838209, 0.509885886451229, 0.467154993209988, 0.579694276675582, 0.638916722033173, 0.820960450684652, 0.714775806758553, 0.744311280548573, 0.564026765991002, 0.362368865637109, 0.462958104442805, 0.102950824191794, 0.555984315695241, 0.212808449054137, 0.105542579898611, 0.184486185899004, 0.366830383660272, 0.91822028066963, 0.325688586570323, 0.758793268585578, 0.387320505687967, 0.124187859939411, 0.838123351102695, 0.566107266349718, 0.856558425119147, 0.309080191655084, 0.0832008349243551, 0.498546092305332, 0.00765023194253445, 0.787346028489992, 0.135992744006217, 0.188882738584653, 0.132122415583581, 0.00905343005433679, 0.967579723568633, 0.548388948198408, 0.251580744748935, 0.425418507773429, 0.304536815499887, 0.0883230513427407, 0.108812490478158, 0.850357534363866, 0.475749591365457, 0.61878112773411, 0.546459384262562, 0.0242443711031228, 0.569645348470658, 0.405287738656625, 0.248911825474352, 0.239915371406823, 0.727432867279276, 0.608208185061812, 0.465208199806511, 0.0426500146277249, 0.619972537271678, 0.997476893942803, 0.0828336265403777, 0.487225137185305, 0.419896790524945, 0.665380371501669, 0.978877859655768, 0.284626831300557, 0.62149839871563, 0.947542705805972, 0.101011359132826, 0.201350795570761, 0.349117855541408, 0.298553706146777, 0.96194491093047, 0.105815237388015, 0.456203067442402, 0.661251310259104, 0.378837346099317, 0.852007883135229, 0.114014406921342, 0.270102264592424, 0.029426303692162, 0.00257087149657309, 0.584946998627856, 0.0210587584879249, 0.497009889921173, 0.334594627842307, 0.311218160204589, 0.123759063659236, 0.964945881860331, 0.69145336584188, 0.502254946157336, 0.343997529242188, 0.89089303323999, 0.175468302099034, 0.0174966906197369, 0.199275871505961, 0.951461914926767, 0.204735714942217, 0.234839039156213, 0.463070431724191, 0.741668940987438, 0.749687527073547, 0.513209325028583, 0.228731067618355, 0.526928115868941, 0.838701426284388, 0.0554709599819034, 0.246625587577, 0.811996770324185, 0.6130228897091, 0.573350707767531, 0.717667723074555, 0.199223013129085, 0.064176837913692, 0.387189425993711, 0.759013240225613, 0.996777210617438, 0.770870657637715, 0.0358003294095397, 0.568873013136908, 0.632267397595569, 0.763558545382693, 0.38213377003558, 0.777259723749012, 0.552731106989086, 0.286718447459862, 0.15803112112917, 0.00151633401401341, 0.707404375774786, 0.959480264224112, 0.338935037143528, 0.909615153912455, 0.84887577733025, 0.255912644788623, 0.147485082736239, 0.570457843597978, 0.966799169313163, 0.19317838922143, 0.375192771432921, 0.213296546135098, 0.142746648285538, 0.102666350547224, 0.363280685152858, 0.827619023621082, 0.061416928190738, 0.0133660680148751, 0.992682837648317, 0.793712365673855, 0.510655459482223, 0.319302571704611, 0.378954589134082, 0.465589468367398, 0.327951384242624, 0.711368726566434, 0.631651958217844, 0.57790394872427, 0.560415137792006, 0.908195280237123, 0.908266257727519, 0.527093342039734, 0.606991047505289, 0.738266244996339, 0.298864332959056, 0.0209947237744927, 0.250639891019091, 0.586740242317319, 0.9288699333556, 0.961406117072329, 0.265082389814779, 0.0771911051124334, 0.532562376465648, 0.649868326960132, 0.591833241283894, 0.627459400333464, 0.52490603341721, 0.320048143854365, 0.31708334502764, 0.0918766343966126, 0.252153050154448, 0.634847228648141, 0.408960634144023, 0.630423122085631, 0.0270396335981786, 0.886820747051388, 0.104907472385094, 0.813101339386776, 0.243835263419896, 0.033443730790168, 0.818625804036856, 0.0725286002270877, 0.709881704766303, 0.133552724029869, 0.905118191847578, 0.542079787934199, 0.0499925015028566, 0.144431116059422, 0.849776636809111, 0.271456729853526, 0.68541242275387, 0.424725402146578, 0.639214080059901, 0.0856713792309165, 0.453780885552987, 0.90485520218499, 0.174402926582843, 0.194436096120626, 0.943772557191551, 0.477028760127723, 0.690638636937365, 0.994714014464989, 0.258037877036259, 0.429787203669548, 0.42454350204207, 0.342918710783124, 0.411770668346435, 0.806799881625921, 0.911958707729355, 0.305239001521841, 0.438101651612669, 0.717707177158445, 0.870280905393884, 0.236212054966018, 0.981998065952212, 0.617577978176996, 0.981360807782039, 0.679342652671039, 0.118014363804832, 0.0529644612688571, 0.0614244984462857, 0.993100913241506, 0.390768539393321, 0.618349893949926, 0.908648143755272, 0.63829514128156, 0.922410153551027, 0.183075831504539), actual = c(0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"