Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>      0.000000000      1.000000000      0.008885619      0.577819889 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>      0.500000000      0.487833252      0.487833252      0.491582487 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.809934912249446, 0.662982622161508, 0.751068250508979, 0.874700480606407, 0.914284062106162, 0.661130387801677, 0.678973474772647, 0.246136462315917, 0.15768410009332, 0.3105232140515, 0.987650366732851, 0.255306048318744, 0.560933247441426, 0.0321792850736529, 0.699637025827542, 0.738722449168563, 0.664705020142719, 0.168192658107728, 0.550707354210317, 0.245719811646268, 0.719412872800604, 0.373222243273631, 0.219701139722019, 0.902264656731859, 0.295812135096639, 0.925222774967551, 0.945909038884565, 0.28024409012869, 0.314750476973131, 0.580311168683693, 0.578683024272323, 0.10232039145194, 0.239935376215726, 0.949321487452835, 0.835269995033741, 0.446777668548748, 0.911470785504207, 0.970208069775254, 0.911923147970811, 0.039758003083989, 0.93941926304251, 0.671552159124985, 0.536122659454122, 0.591597834136337, 0.530628527514637, 0.777660753112286, 0.410690025659278, 0.229743250878528, 0.724554181331769, 0.2164605865255, 0.887819565832615, 0.60493948450312, 0.716883469372988, 0.34284590696916, 0.855967074399814, 0.21968910517171, 0.341328653506935, 0.588713939767331, 0.753629493061453, 0.527398766251281, 0.491669752635062, 0.724255999550223, 0.421185551444069, 0.440525222802535, 0.288376048440114, 0.76124724955298, 0.559451278764755, 0.486404301831499, 0.383810399100184, 0.917990466579795, 0.730533347232267, 0.662687283009291, 0.172888196073472, 0.577658123802394, 0.839565406786278, 0.188111235620454, 0.930283651454374, 0.66389759676531, 0.931776299839839, 0.573710046941414, 0.0096554565243423, 0.857692325254902, 0.159576810430735, 0.573478125967085, 0.898041643202305, 0.783466070424765, 0.510501171462238, 0.25660774554126, 0.371321931947023, 0.789267963962629, 0.209784721955657, 0.497224457096308, 0.108056815108284, 0.13628077856265, 0.78950998862274, 0.104536842787638, 0.0903817259240896, 0.185756515944377, 0.772038198076189, 0.3589955244679, 0.0944011679384857, 0.872801472432911, 0.287890643114224, 0.268516173120588, 0.684495513094589, 0.630348403938115, 0.701391012873501, 0.313966829096898, 0.619604514213279, 0.519131416454911, 0.940730106784031, 0.106038556667045, 0.0890263996552676, 0.497188102453947, 0.164695963729173, 0.99526892718859, 0.652141028083861, 0.224463933380321, 0.470965836429968, 0.967671345220879, 0.877196792047471, 0.402107662055641, 0.233521547634155, 0.665358338970691, 0.933012898545712, 0.816404961515218, 0.722032297868282, 0.680632721399888, 0.165629676543176, 0.0190835138782859, 0.0843458047602326, 0.976573494262993, 0.342438535299152, 0.241734228096902, 0.962345537962392, 0.750565598253161, 0.903589141089469, 0.807927573798224, 0.371720237191767, 0.652846432290971, 0.384625698206946, 0.22267557727173, 0.626563359983265, 0.410525093087927, 0.440962062450126, 0.551777636865154, 0.759576957672834, 0.172402611467987, 0.869439996778965, 0.0838013705797493, 0.631398870144039, 0.151053370442241, 0.826569992117584, 0.0473506096750498, 0.345643857028335, 0.374283920973539, 0.00745827471837401, 0.284273562720045, 0.882420880487189, 0.184597202809528, 0.239188632229343, 0.954491295618936, 0.515928049338982, 0.553995256545022, 0.413718591677025, 0.467571303714067, 0.0638778025750071, 0.397893394110724, 0.704288307577372, 0.744842938845977, 0.912264412268996, 0.717416354687884, 0.189533262513578, 0.0528046805411577, 0.621341315563768, 0.362767489627004, 0.63888994907029, 0.981235765386373, 0.996931875590235, 0.841530062491074, 0.99576160358265, 0.887660905485973, 0.595176184317097, 0.397496273042634, 0.907624258194119, 0.0720623550005257, 0.805590502452105, 0.576171290827915, 0.784225595416501, 0.0331773061770946, 0.506294792285189, 0.89424353512004, 0.887839620234445, 0.340773786185309, 0.580216629430652, 0.505513621959835, 0.422087892191485, 0.364537809276953, 0.396887930575758, 0.0146426835563034, 0.616759314900264, 0.232525342144072, 0.372090010205284, 0.765627044951543, 0.345871759345755, 0.104948034510016, 0.949965872336179, 0.698826208245009, 0.127261901739985, 0.476710814516991, 0.333491401048377, 0.0602480359375477, 0.622139114420861, 0.265286872396246, 0.606251047225669, 0.771679991856217, 0.544485834427178, 0.101242597447708, 0.590662298491225, 0.454867624444887, 0.119990598876029, 0.836022712057456, 0.710583139210939, 0.128520737402141, 0.0249384914059192, 0.818382610101253, 0.356239303480834, 0.356564634013921, 0.519401547499001, 0.805301891174167, 0.0858188690617681, 0.645206753630191, 0.377629795344546, 0.886774372542277, 0.912730435607955, 0.583829943556339, 0.983884677058086, 0.539006491191685, 0.723082366865128, 0.68349846219644, 0.150940854800865, 0.513886940432712, 0.475083919009194, 0.880793956108391, 0.173312193248421, 0.150503832614049, 0.773930327966809, 0.793373462511227, 0.97136539593339, 0.303254817612469, 0.71374565991573, 0.236403961433098, 0.730180665152147, 0.0511181319598109, 0.458879164187238, 0.295705741038546, 0.360733634326607, 0.844962806208059, 0.709500576835126, 0.0881664417684078, 0.419725304702297, 0.270972465863451, 0.52843233011663, 0.0487414246890694, 0.236494874348864, 0.947249558987096, 0.959636476123706, 0.676340395119041, 0.591724808793515, 0.757170089520514, 0.384145837044343, 0.189242880558595, 0.33123653777875, 0.000905857654288411, 0.712218078784645, 0.872196155367419, 0.082038787426427, 0.411252212943509, 0.538441385375336, 0.885421226965263, 0.32744274311699, 0.871150978375226, 0.809211052022874, 0.167770517058671, 0.903236892074347, 0.689492502249777, 0.376796344295144, 0.614671368384734, 0.671677052974701, 0.137515676440671, 0.324075954966247, 0.612061344785616, 0.99402691423893, 0.768311444204301, 0.603984939632937, 0.250989191699773, 0.0316191385500133, 0.373737516114488, 0.264935518847778, 0.0967251725960523, 0.553638492710888, 0.573581061325967, 0.086719183716923, 0.671243828954175, 0.0191664656158537, 0.285728433402255, 0.82158705801703, 0.166346137179062, 0.29635341023095, 0.950828633271158, 0.116004487732425, 0.988558470737189, 0.609601792180911, 0.402042655274272, 0.193417998729274, 0.759693530155346, 0.0829186767805368, 0.865395701956004, 0.0387175476644188, 0.204605820821598, 0.00970154884271324, 0.885694215074182, 0.774427911732346, 0.294811999890953, 0.0965907436329871, 0.064994775922969, 0.278790953569114, 0.0634515243582428, 0.228336566826329, 0.0227225772105157, 0.406314788386226, 0.469053151784465, 0.351323499111459, 0.43128838040866, 0.0534903577063233, 0.309393303235993, 0.350813307100907, 0.984724717913195, 0.489677139325067, 0.38771679578349, 0.839485291624442, 0.590443936176598, 0.796534071210772, 0.690961740212515, 0.03843406541273, 0.806865531951189, 0.987328606890514, 0.0739704484585673, 0.2070544117596, 0.386565715074539, 0.292065284214914, 0.806573024950922, 0.946854123612866, 0.210859728278592, 0.914096190128475, 0.896191664971411, 0.673655265243724, 0.0673375725746155, 0.147556134033948, 0.548185271443799, 0.325167134869844, 0.370072045363486, 0.535791437840089, 0.489584833849221, 0.937767133349553, 0.827640869189054, 0.0774468218442053, 0.16117883566767, 0.65143706346862, 0.0560229765251279, 0.50215101102367, 0.00921270321123302, 0.664791303919628, 0.541125781834126, 0.915002349764109, 0.842746405629441, 0.0973845974076539, 0.485607143957168, 0.891931798541918, 0.402149844914675, 0.560262084705755, 0.104680731659755, 0.0245109279640019, 0.281140147475526, 0.303118439391255, 0.770877829287201, 0.00130485370755196, 0.167390659218654, 0.52903740410693, 0.320392570458353, 0.553033750038594, 0.0807199655100703, 0.674750727135688, 0.475287470500916, 0.657630478963256, 0.358495341381058, 0.913668775465339, 0.917844660580158, 0.782339857425541, 0.129899803549051, 0.129094499628991, 0.014800813049078, 0.166465176269412, 0.0265529449097812, 0.633189230924472, 0.349948603892699, 0.326124908169731, 0.707430053502321, 0.869566408451647, 0.698079428868368, 0.374022281728685, 0.871099045267329, 0.635276763001457, 0.0692604414653033, 0.988764027599245, 0.432377347024158, 0.966091738315299, 0.311374575132504, 0.843222230672836, 0.190349563024938, 0.462890124879777, 0.114062598673627, 0.0166301671415567, 0.617907012347132, 0.763548162765801, 0.0918240312021226, 0.439368575345725, 0.496212068246678, 0.149749876698479, 0.472014846513048, 0.908392954384908, 0.502063722116873, 0.113313481677324, 0.775744878919795, 0.37912790547125, 0.336685581831262, 0.124135368037969, 0.431505963439122, 0.713091070996597, 0.238667418016121, 0.507507014553994, 0.451661994680762, 0.190830414416268, 0.308213244890794, 0.467446628957987, 0.166154893347993, 0.235733170993626, 0.939629088155925, 0.873332373797894, 0.0828459651675075, 0.414136650972068, 0.85283614625223, 0.428093277383596, 0.552980014588684, 0.913280017208308, 0.705406935885549, 0.829245316796005, 0.65704503073357, 0.358718009432778, 0.768565263831988, 0.276601094054058, 0.546552925836295, 0.632117927772924, 0.363806257257238, 0.791837830562145, 0.27271962701343, 0.377694514812902, 0.32528336532414, 0.783047300297767, 0.575561132980511, 0.351807844825089, 0.150198055431247, 0.746143993455917, 0.362172024324536, 0.0144522555638105, 0.540133164264262, 0.152327596908435, 0.543546220520511, 0.873285048874095, 0.852593931136653, 0.749940378824249, 0.151065883226693, 0.359911165665835, 0.757730033714324, 0.727851270698011, 0.422472572419792, 0.65103535912931, 0.510877262568101, 0.251261411700398, 0.553737253881991, 0.651489460607991, 0.678219955647364, 0.922112347790971, 0.722598711727187, 0.697169356746599, 0.403182960348204, 0.402818540576845, 0.833089074818417, 0.782805775525048, 0.970249415608123, 0.276250886265188, 0.936753088375553, 0.811331069096923, 0.76080190972425, 0.170161242363974, 0.182431822875515, 0.82567851874046, 0.296087017515674, 0.794468525098637, 0.100565880071372, 0.856684413738549, 0.166734782513231, 0.258592024445534, 0.95723115070723, 0.333757241256535, 0.166122365510091, 0.337753827916458, 0.922862692736089, 0.650894924299791, 0.234774457989261, 0.388937537558377, 0.864889171672985, 0.245647495612502, 0.14813685673289, 0.343481270596385, 0.171048273798078, 0.202467154944316, 0.810348931699991, 0.296175624011084, 0.155751348473132, 0.536954000359401, 0.628835693234578, 0.0482886312529445, 0.122757733799517, 0.955732860369608, 0.977981701726094, 0.38650551927276, 0.624614606844261, 0.94395279022865, 0.22008337941952, 0.00577658228576183, 0.642160386079922, 0.794255231507123, 0.247650708770379, 0.664411864476278, 0.36591938091442, 0.265125988516957, 0.559942715801299, 0.336410726886243, 0.353190527064726, 0.270121597452089, 0.616041913628578, 0.451945899985731, 0.144279332598671, 0.255919875809923, 0.284022470936179, 0.268461557803676, 0.901739018503577, 0.340616799658164, 0.0379622760228813, 0.223853084025905, 0.0569648453965783, 0.162602422060445, 0.288397297263145, 0.391238951357082, 0.309998288284987, 0.258851166814566, 0.8542762727011, 0.631736520677805, 0.437854415038601, 0.850033569615334, 0.810045470716432, 0.933573949383572, 0.18599611450918, 0.381127365631983, 0.607942717382684, 0.434586032759398, 0.158464308828115, 0.305412835907191, 0.0194200084079057, 0.779165871674195, 0.593865222297609, 0.181712182471529, 0.237054363824427, 0.954917411552742, 0.79955224134028, 0.550721548497677, 0.461165055632591, 0.672890300862491, 0.327566062100232, 0.962127882288769, 0.561880592722446, 0.0249393812846392, 0.473216140642762, 0.300521355122328, 0.174101943382993, 0.676065652165562, 0.430680562276393, 0.711606541648507, 0.908046143129468, 0.225381610216573, 0.151238118065521, 0.195128375664353, 0.441329101333395, 0.609439289197326, 0.555521590868011, 0.82203385210596, 0.449978440301493, 0.933937087655067, 0.655780666507781, 0.947995662689209, 0.522414726670831, 0.893165276153013, 0.177263017278165, 0.442084396257997, 0.0679594986140728, 0.464502663584426, 0.777869266457856, 0.799606362357736, 0.476356772705913, 0.974036716856062, 0.393284798366949, 0.826138578820974, 0.933177080005407, 0.846889377571642, 0.0292743693571538, 0.863886270206422, 0.0862215366214514, 0.675678994972259, 0.567802630597726, 0.152115932898596, 0.764916350366548, 0.921082626096904, 0.828483211342245, 0.921677579171956, 0.715546371415257, 0.298736112192273, 0.509351596236229, 0.279247100930661, 0.127807522425428, 0.969032296445221, 0.0336968759074807, 0.0162898865528405, 0.611420892411843, 0.980745719745755, 0.715898594819009, 0.859756294405088, 0.113356843125075, 0.505949303740636, 0.0906627534423023, 0.554668825352564, 0.871452698716894, 0.128747192909941, 0.413262937916443, 0.975749244447798, 0.43698455975391, 0.813235404435545, 0.513357481919229, 0.31978415697813, 0.189010690664873, 0.929837473668158, 0.846524736378342, 0.328304841415957, 0.338621528120711, 0.406188101973385, 0.935121755348518, 0.889825348509476, 0.883354274788871, 0.470614387420937, 0.513514247024432, 0.541764747817069, 0.133845553966239, 0.976276961620897, 0.0653550524730235, 0.329206197988242, 0.624814773211256, 0.618102073203772, 0.415226685814559, 0.712049764813855, 0.3385873620864, 0.778156677028164, 0.123602851293981, 0.601694003678858, 0.899543100269511, 0.851352339377627, 0.0626441536005586, 0.119480482535437, 0.0808661598712206, 0.660457689547911, 0.528369380393997, 0.098880864912644, 0.607529679778963, 0.546563959680498, 0.368939624167979, 0.63872288656421, 0.860844505950809, 0.878146999981254, 0.805232015205547, 0.652379381237552, 0.216022017644718, 0.26375511335209, 0.158078293083236, 0.0836474404204637, 0.74757649214007, 0.42652437556535, 0.210195207968354, 0.712019586935639, 0.090769256465137, 0.586752516683191, 0.741530659142882, 0.647865536622703, 0.72758892015554, 0.0930655824486166, 0.0189303883817047, 0.839305686764419, 0.280233351280913, 0.198653016239405, 0.248334004543722, 0.599196079885587, 0.271681687328964, 0.0272397720254958, 0.67204678338021, 0.14239630359225, 0.259269956964999, 0.381075167795643, 0.871539711253718, 0.404741201316938, 0.775402838829905, 0.730682007735595, 0.747721468331292, 0.86972818034701, 0.363659530645236, 0.479845890775323, 0.501381669193506, 0.188593670027331, 0.615988653386012, 0.246139058843255, 0.870711085153744, 0.375781797571108, 0.929262524703518, 0.923944126814604, 0.416032700566575, 0.816931338747963, 0.796969558577985, 0.644471826264635, 0.394050993956625, 0.673237508162856, 0.520847037434578, 0.76633316394873, 0.492184140952304, 0.831801689229906, 0.66787571227178, 0.328550077974796, 0.384949397528544, 0.816187605960295, 0.713686966337264, 0.00452648894861341, 0.312645096564665, 0.0669918763451278, 0.865802452201024, 0.931186119094491, 0.303325777640566, 0.306358526926488, 0.959254145389423, 0.158872795756906, 0.22715652664192, 0.707362564979121, 0.810859917895868, 0.0896961989346892, 0.77080305153504, 0.939960626652464, 0.837581308791414, 0.655762915266678, 0.655679692048579, 0.418870084686205, 0.737267842981964, 0.116059169871733, 0.0556475652847439, 0.103229423053563, 0.607387451687828, 0.513511628378183, 0.242377922870219, 0.0235953808296472, 0.273002391215414, 0.306369286961854, 0.643861654447392, 0.71055523538962, 0.946808569598943, 0.746717554749921, 0.830049429321662, 0.227254461497068, 0.185028461972252, 0.281548566650599, 0.622275564586744, 0.328684899723157, 0.652174870716408, 0.567512702895328, 0.505462238565087, 0.41141813993454, 0.1792089745868, 0.870839623268694, 0.608822711277753, 0.618754494003952, 0.223263349849731, 0.336606824304909, 0.81839841697365, 0.0220160728786141, 0.671286410652101, 0.186855196719989, 0.236465121386573, 0.920188783900812, 0.541226465487853, 0.163781066425145, 0.0645130011253059, 0.677154641831294, 0.400715617928654, 0.0618251890409738, 0.601096280384809, 0.827702158829197, 0.99675815878436, 0.805643710307777, 0.353649072116241, 0.870489520719275, 0.285186774330214, 0.777972894255072, 0.662860503885895, 0.608322595478967, 0.196906196186319, 0.727263304172084, 0.908438759157434, 0.943159418180585, 0.455214624525979, 0.870064419694245, 0.123611437622458, 0.562461989698932, 0.954807570204139, 0.253271477529779, 0.0943855447694659, 0.910805886145681, 0.285690947668627, 0.892512031598017, 0.870404633227736, 0.644963866099715, 0.505844380473718, 0.633876584935933, 0.330489083658904, 0.703930393559858, 0.760986339300871, 0.549393359804526, 0.443928041495383, 0.150925410911441, 0.503968675155193, 0.193170714657754, 0.676731177605689, 0.81840031221509, 0.891869237180799, 0.897519690450281, 0.842543386854231, 0.504063458181918, 0.304753285599872, 0.119649451691657, 0.300027355318889, 0.79222174291499, 0.991336736828089, 0.816862193401903, 0.736765663838014, 0.253420466789976, 0.410904757911339, 0.404619080713019, 0.000394649337977171, 0.168485290138051, 0.550873643020168, 0.801315708784387, 0.114773886976764, 0.688277738634497, 0.481164373690262, 0.72113748639822, 0.437452910933644, 0.924433093750849, 0.229772238060832, 0.811853860737756, 0.636700470233336, 0.526335306931287, 0.0926402509212494, 0.21246775961481, 0.212175864260644, 0.478928054217249, 0.745387794449925, 0.761220245622098, 0.444293101318181, 0.804216088727117, 0.969818038865924, 0.260297280736268, 0.739234906388447, 0.0562969285529107, 0.38973277551122, 0.316493302816525, 0.0339857020881027, 0.788231442449614, 0.28287063841708, 0.0571843313518912, 0.731247912859544, 0.452668437501416, 0.127972536487505, 0.357819646829739, 0.427759434096515, 0.991973077179864, 0.718303353525698, 0.0955516719259322, 0.990735023515299, 0.920430992962793, 0.0864748554304242, 0.176522077061236, 0.792951685842127, 0.854258682345971, 0.958785661961883, 0.573470145929605, 0.127539003733546, 0.515305267646909, 0.894223917275667, 0.517023263033479, 0.435208658222109, 0.508752536028624, 0.981935693183914, 0.318222841946408, 0.85895779938437, 0.242377191083506, 0.942559440154582, 0.436615925515071, 0.523398624267429, 0.413596211699769, 0.939849947346374, 0.374010741245002, 0.0275967658963054, 0.894014865159988, 0.488225131528452, 0.94861757196486, 0.685041833901778, 0.212281227111816, 0.57903426582925, 0.387077119899914, 0.971859315643087, 0.92017470067367, 0.518978784559295, 0.396819576155394, 0.744402579497546, 0.660212239949033, 0.503320067655295, 0.573631368111819, 0.242387963458896, 0.603104644687846, 0.319859639042988, 0.107629904523492, 0.780078453943133, 0.320190293947235, 0.358349803835154, 0.779561029514298, 0.10652718716301, 0.931601146701723, 0.507634880254045, 0.127576770959422, 0.548569488571957, 0.966627232497558, 0.506363979540765, 0.124030032195151, 0.204315505223349, 0.942327580647543, 0.555223846109584, 0.42809398798272, 0.00736060738563538, 0.961732293944806, 0.983097597723827, 0.366200785385445, 0.562991213053465, 0.289491443429142, 0.216459396062419, 0.357412258395925, 0.0821485361084342, 0.631705847335979, 0.0555067860987037, 0.254912653239444, 0.59287386550568, 0.695365119725466, 0.562970707425848, 0.10176840913482, 0.537897063652053, 0.675371107412502, 0.109227248933166, 0.498374461894855, 0.0439803507179022, 0.221664700191468, 0.628970288904384, 0.674761353759095, 0.356684845173731, 0.180118520976976, 0.879561190726236, 0.422442160081118, 0.616982321254909, 0.785337424604222, 0.510830385610461), actual = c(1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.5845915467944, 0.264958241488785, 0.674273398239166, 0.334167017834261, 0.915954618016258, 0.355693307006732, 0.0431060993578285, 0.855013203341514, 0.517862202366814, 0.945286824833602, 0.222629673313349, 0.887979067396373, 0.367984529118985, 0.893555752234533, 0.0512798177078366, 0.28649184294045, 0.257467185845599, 0.0888208465185016, 0.252353768795729, 0.841692902613431, 0.424849083879963, 0.863852703711018, 0.297419091453776, 0.00277915573678911, 0.647120325360447, 0.00223063048906624, 0.412825838429853, 0.0264358518179506, 0.700133653357625, 0.676367848645896, 0.292642368236557, 0.0769264944829047, 0.106568542076275, 0.884281689766794, 0.769456909503788, 0.703814664622769, 0.763540316838771, 0.463950487552211, 0.569628872442991, 0.0187403454910964, 0.932993621798232, 0.544803526252508, 0.282199649838731, 0.23610373865813, 0.921773299342021, 0.687188351526856, 0.497528470586985, 0.559099563397467, 0.236167734721676, 0.859145464608446, 0.260567639255896, 0.683751723961905, 0.852189793949947, 0.949442688841373, 0.713729345938191, 0.102997351437807, 0.0409039142541587, 0.367615255294368, 0.889224624726921, 0.945827731164172, 0.550699958344921, 0.974499886389822, 0.250066048232839, 0.0837967910338193, 0.102896814700216, 0.201513253385201, 0.716783120529726, 0.330147066852078, 0.441692364867777, 0.29498388338834, 0.136963941389695, 0.914069464430213, 0.0668177390471101, 0.307432169094682, 0.378967557800934, 0.374597756657749, 0.955056291073561, 0.81564224883914, 0.0321808806620538, 0.0187891114037484, 0.169826337136328, 0.420708552468568, 0.558389796875417, 0.999260851182044, 0.00464627170003951, 0.151345159392804, 0.743769548134878, 0.183191983262077, 0.308050479507074, 0.89632560685277, 0.799133149906993, 0.712465634569526, 0.0384887843392789, 0.558162262896076, 0.67968930141069, 0.266483844025061, 0.256225458346307, 0.17614819156006, 0.359126148978248, 0.643555442336947, 0.657872164389119, 0.344311948399991, 0.994533082004637, 0.919592957943678, 0.372624346986413, 0.636401033727452, 0.0686418579425663, 0.973971841624007, 0.514787787571549, 0.369032123591751, 0.972764310892671, 0.980343982111663, 0.457540998235345, 0.670882634120062, 0.159154095454141, 0.932828823802993, 0.322787339799106, 0.14875725377351, 0.225547155365348, 0.530059681041166, 0.13229798222892, 0.0377875918056816, 0.809990399051458, 0.774360365001485, 0.599737515905872, 0.832516981754452, 0.106163116171956, 0.135339963482693, 0.130708063719794, 0.300339401932433, 0.870792594505474, 0.696697976905853, 0.686779885319993, 0.405540849082172, 0.211576177505776, 0.840733416844159, 0.613958119414747, 0.0315328389406204, 0.696638070046902, 0.462892406852916, 0.0390053323935717, 0.992998465895653, 0.919433229137212, 0.234306848840788, 0.672068758169189, 0.307967714266852, 0.323262311285362, 0.913596107624471, 0.275751168606803, 0.564348810352385, 0.155798828694969, 0.142707016319036, 0.164395522326231, 0.0864950821269304, 0.203039200277999, 0.335626863408834, 0.945144358091056, 0.557365056592971, 0.150957313831896, 0.188350511249155, 0.272185715846717, 0.865329627413303, 0.718270333949476, 0.0940519543364644, 0.812118658795953, 0.50122642936185, 0.256714549381286, 0.171198369003832, 0.102935113245621, 0.385050543118268, 0.00299089122563601, 0.0512013945262879, 0.681274573551491, 0.808536070631817, 0.409263128647581, 0.208121174015105, 0.605693759163842, 0.878250902285799, 0.117051335750148, 0.488209635717794, 0.913382740225643, 0.875543126137927, 0.00281408615410328, 0.78980762604624, 0.879835153697059, 0.443992301123217, 0.696479508886114, 0.79349401104264, 0.677334909792989, 0.687814695294946, 0.615075291134417, 0.863428493728861, 0.56888888636604, 0.739712591050193, 0.0393825198989362, 0.651960659073666, 0.933545222505927, 0.657254952006042, 0.0472403126768768, 0.407979118172079, 0.311709910398349, 0.423832177417353, 0.962155542336404, 0.410838999552652, 0.840001924661919, 0.902535484172404, 0.916328026680276, 0.33209337037988, 0.873403782956302, 0.168300574412569, 0.0491430370602757, 0.837994182249531, 0.334036642452702, 0.13589949416928, 0.632963789626956, 0.539841333171353, 0.277240892639384, 0.11579127702862, 0.748487380566075, 0.68542745988816, 0.975263083353639, 0.27824427979067, 0.794439778197557, 0.589404644677415, 0.757712046150118, 0.196487287757918, 0.464599568396807, 0.319691397948191, 0.0218919853214175, 0.41307467664592, 0.21733105997555, 0.61990576935932, 0.627846414223313, 0.757376891328022, 0.126802972517908, 0.677723693428561, 0.522501095430925, 0.477229089941829, 0.486605498706922, 0.601795900147408, 0.0815675973426551, 0.789486594963819, 0.446976082166657, 0.907275529345497, 0.69088838249445, 0.169477440416813, 0.451058665756136, 0.285735405515879, 0.154505466111004, 0.410007117781788, 0.961372513324022, 0.841264080954716, 0.687779888976365, 0.555472612846643, 0.00390959507785738, 0.851086530601606, 0.961003966629505, 0.978990513365716, 0.435742665315047, 0.484630565857515, 0.4798504258506, 0.875714426161721, 0.169698096578941, 0.319987905677408, 0.0920465069357306, 0.413143638521433, 0.539228057488799, 0.684943189611658, 0.473626147257164, 0.997327475110069, 0.662154135061428, 0.436104470165446, 0.863385057076812, 0.651974382111803, 0.865492933429778, 0.237460907083005, 0.913458868395537, 0.896491902880371, 0.77362469304353, 0.111100973561406, 0.813237515976653, 0.685000736732036, 0.429169309325516, 0.460498726228252, 0.485765937017277, 0.925322365947068, 0.419208823237568, 0.854667134350166, 0.130447677103803, 0.417308523319662, 0.423131908290088, 0.448881370015442, 0.944800146156922, 0.270280242664739, 0.025593192782253, 0.117403974290937, 0.857941293390468, 0.34394779196009, 0.629060038598254, 0.750542908208445, 0.886671688873321, 0.7201628445182, 0.0809473877307028, 0.428969954838976, 0.336046736687422, 0.826657991623506, 0.711604576557875, 0.379595924401656, 0.0252216241788119, 0.830886184936389, 0.672766691073775, 0.786932385293767, 0.0356068548280746, 0.861623759847134, 0.464061988284811, 0.206468735588714, 0.895053849788383, 0.220946869114414, 0.723983484786004, 0.390197044238448, 0.0958386263810098, 0.853233569534495, 0.637686101254076, 0.734436046797782, 0.506104127271101, 0.733402119483799, 0.802247384097427, 0.894357849145308, 0.257772700162604, 0.0456095424015075, 0.604160938877612, 0.102966071339324, 0.405518634943292, 0.0170047844294459, 0.0692644186783582, 0.0614797824528068, 0.222540306625888, 0.504790546139702, 0.76930784014985, 0.517828979995102, 0.713717028032988, 0.847714787581936, 0.317755218362436, 0.358878501923755, 0.189710367936641, 0.757840130943805, 0.950247788801789, 0.503905362216756, 0.883683766704053, 0.693028395762667, 0.899179243482649, 0.492846592329443, 0.304102080408484, 0.406843091826886, 0.651923868339509, 0.968170856824145, 0.0294510766398162, 0.529206593055278, 0.298463076818734, 0.0598771227523685, 0.653434352949262, 0.0338700891006738, 0.0503874949645251, 0.788295199396089, 0.361114817438647, 0.513629744760692, 0.709052172256634, 0.0238359966315329, 0.294772367691621, 0.691369589883834, 0.579780848929659, 0.252440667478368, 0.0979723113123327, 0.227399638853967, 0.490714172599837, 0.348453801125288, 0.383585997391492, 0.462059352779761, 0.464646168285981, 0.282325866632164, 0.859758149133995, 0.216838296502829, 0.993658056249842, 0.200836329720914, 0.539155701175332, 0.973509328905493, 0.721475630067289, 0.0457188771106303, 0.657015998149291, 0.440589450998232, 0.464690506225452, 0.344346263911575, 0.778345673345029, 0.831384160323068, 0.889855898916721, 0.565038925036788, 0.45720523246564, 0.227202529786155, 0.0678015314042568, 0.104178921785206, 0.940464029088616, 0.658297001617029, 0.8343151723966, 0.711210949113593, 0.326872754376382, 0.177437061676756, 0.609515772433951, 0.233925189822912, 0.949374523945153, 0.72346026613377, 0.635015510488302, 0.0792198709677905, 0.405560145387426, 0.986197653226554, 0.92274396494031, 0.643500100122765, 0.695390035398304, 0.847866285359487, 0.202725903363898, 0.702363519463688, 0.791379832429811, 0.310188860632479, 0.94853847916238, 0.993252288782969, 0.934163229074329, 0.455275358399376, 0.519770427374169, 0.309773420915008, 0.481479465961456, 0.72347422875464, 0.728389016818255, 0.298130589770153, 0.7366043694783, 0.951825376600027, 0.780088901286945, 0.863197448197752, 0.991498263087124, 0.912966606440023, 0.978802463039756, 0.365216226549819, 0.447624871972948, 0.221040226984769, 0.882318514864892, 0.482533945469186, 0.431322782533243, 0.573408990865573, 0.83007293054834, 0.145164174260572, 0.268571741646156, 0.216337333200499, 0.421866463962942, 0.300901161739603, 0.881143618142232, 0.456159751629457, 0.263720533112064, 0.890876506688073, 0.956932202680036, 0.718149851076305, 0.249383767601103, 0.84121468779631, 0.356110418913886, 0.0777933609206229, 0.344433559803292, 0.288450191728771, 0.981682698475197, 0.602831884054467, 0.23467633058317, 0.616485398961231, 0.546688787639141, 0.0132985503878444, 0.12495263828896, 0.569557403214276, 0.788082395680249, 0.443915934301913, 0.584656502353027, 0.148808755213395, 0.895502823870629, 0.307474141474813, 0.0316190358716995, 0.457079579820856, 0.596524843014777, 0.608058477053419, 0.758858382003382, 0.742754457518458, 0.317572828847915, 0.953898763982579, 0.31299330922775, 0.950068406527862, 0.154568304074928, 0.0214967490173876, 0.147456675302237, 0.0200889313127846, 0.39573186612688, 0.841887543676421, 0.431783393258229, 0.335836795624346, 0.577630317304283, 0.369513956131414, 0.919337473344058, 0.939795992569998, 0.00535419001244009, 0.123067963402718, 0.756782902870327, 0.444249885389581, 0.732017337810248, 0.744423354743049, 0.900479908334091, 0.0320051198359579, 0.204328344669193, 0.0624235309660435, 0.492343675345182, 0.501888379221782, 0.163181392708793, 0.0699551147408783, 0.080056338571012, 0.744073488283902, 0.939188396325335, 0.649104641051963, 0.390303868567571, 0.673649844713509, 0.96016103005968, 0.173097747610882, 0.973188974661753, 0.232020649127662, 0.753756301710382, 0.661039020633325, 0.647088069003075, 0.210070681292564, 0.63034162693657, 0.0622909285593778, 0.426698233932257, 0.49983979947865, 0.805714435642585, 0.101958272745833, 0.0290110891219229, 0.374408702831715, 0.882379126735032, 0.7507126966957, 0.44067335431464, 0.356862171320245, 0.344992795027792, 0.356419390300289, 0.613580527016893, 0.824142477475107, 0.31572494469583, 0.618369763251394, 0.991565017029643, 0.0963392576668411, 0.895632372936234, 0.69374322029762, 0.0168902475852519, 0.849456172669306, 0.953338617924601, 0.901007732842118, 0.30936021101661, 0.0744752585887909, 0.245976842241362, 0.0297139585018158, 0.967050417326391, 0.633604290895164, 0.288581900997087, 0.147772101452574, 0.260148044209927, 0.602275278419256, 0.658837891183794, 0.826590820448473, 0.146085577551275, 0.224337105639279, 0.341895457357168, 0.238457107450813, 0.580894431564957, 0.0792430972214788, 0.853458050638437, 0.0279823788441718, 0.391306560020894, 0.39100118377246, 0.736085209762678, 0.890995837748051, 0.0344521480146796, 0.809393212664872, 0.17599708517082, 0.89296467602253, 0.125880664447322, 0.0274095360655338, 0.432453294983134, 0.172824553679675, 0.983050054404885, 0.466512975282967, 0.449795817956328, 0.503557137213647, 0.357419576263055, 0.360020910622552, 0.400514031760395, 0.867241693893448, 0.869799485662952, 0.48913680552505, 0.189419765025377, 0.686007683863863, 0.581159635912627, 0.197602381231263, 0.729763799346983, 0.599807032151148, 0.155610149726272, 0.38779825437814, 0.248099037911743, 0.866871236357838, 0.0435032658278942, 0.393753812648356, 0.515933881513774, 0.533564330311492, 0.908359319670126, 0.0488374743144959, 0.0746086549479514, 0.886407115031034, 0.598312533460557, 0.39169524400495, 0.818717555142939, 0.894782012095675, 0.634455480612814, 0.990900259232149, 0.485789402388036, 0.0280246192123741, 0.781951446086168, 0.6174904294312, 0.785556914051995, 0.814809925854206, 0.517728604609147, 0.223030322231352, 0.00102078076452017, 0.500797077547759, 0.736701983958483, 0.234844353748485, 0.537326423451304, 0.228085236158222, 0.645327994832769, 0.618194630136713, 0.0527277595829219, 0.570668141590431, 0.386575201526284, 0.105795572977513, 0.301169321639463, 0.328357986873016, 0.840086042648181, 0.544493582099676, 0.632056618109345, 0.457204912090674, 0.188861680682749, 0.111071457387879, 0.758120544720441, 0.1294243526645, 0.228345370385796, 0.960269723087549, 0.206436025211588, 0.405041908146814, 0.96245082304813, 0.930932021932676, 0.69064956298098, 0.209936428815126, 0.396478999406099, 0.358896882738918, 0.271824293304235, 0.264274124288931, 0.000870556337758899, 0.120277917245403, 0.831555380718783, 0.900712261442095, 0.194423666689545, 0.621072923997417, 0.602204474853352, 0.0994448533747345, 0.490989150712267, 0.257191529730335, 0.225466568022966, 0.674154354026541, 0.598894634284079, 0.782919907709584, 0.193038881057873, 0.232028546044603, 0.892274035373703, 0.583995722932741, 0.251288668252528, 0.131141777383164, 0.419990073656663, 0.486897874856368, 0.427104661939666, 0.552758652251214, 0.804558986565098, 0.216407399624586, 0.340206664288417, 0.517263046232983, 0.165454612346366, 0.401108631165698, 0.682559108827263, 0.267590800998732, 0.298098487313837, 0.871720011811703, 0.762124706758186, 0.964979511452839, 0.312190761091188, 0.477595805190504, 0.653188029304147, 0.646953935967758, 0.637855431064963, 0.569688203046098, 0.578638087259606, 0.187271497910842, 0.451998924836516, 0.0175976022146642, 0.466774999164045, 0.206634921720251, 0.422972857952118, 0.140844452660531, 0.479387759929523, 0.336936154402792, 0.806644475786015, 0.468152741203085, 0.507149589713663, 0.0375350664835423, 0.965755318757147, 0.588088861200958, 0.63685122365132, 0.132954517612234, 0.909870117437094, 0.376585456077009, 0.116628443356603, 0.0804275339469314, 0.679581773234531, 0.706344201229513, 0.72809805884026, 0.31668639392592, 0.33381535904482, 0.958289993694052, 0.148635989287868, 0.470683306688443, 0.0283124754205346, 0.745332529302686, 0.45428968523629, 0.187471690354869, 0.13169157714583, 0.000462065683677793, 0.229683165205643, 0.956779123051092, 0.413092351285741, 0.666826947126538, 0.844840491423383, 0.884610562352464, 0.809772994369268, 0.619101229589432, 0.933182408800349, 0.130183791508898, 0.94744676630944, 0.150548386620358, 0.403451793827116, 0.85536992480047, 0.970265100011602, 0.693251842167228, 0.885178179014474, 0.461745838169008, 0.964048799360171, 0.446251202840358, 0.535706211579964, 0.0752669400535524, 0.586302397307009, 0.207895773462951, 0.938217488816008, 0.635822315933183, 0.602883937070146, 0.496236514067277, 0.124625301687047, 0.324987759348005, 0.447464386466891, 0.583891024114564, 0.760135834803805, 0.696801777230576, 0.527895631734282, 0.17496391851455, 0.883787686238065, 0.398287566378713, 0.948054702254012, 0.13213857379742, 0.993925929069519, 0.959232351509854, 0.372691961470991, 0.474486190127209, 0.505951672326773, 0.621193267405033, 0.780096446396783, 0.707529437262565, 0.356362284626812, 0.229943050071597, 0.559865852585062, 0.929204707266763, 0.260085967369378, 0.330826536752284, 0.710866866167635, 0.771285854047164, 0.291324120247737, 0.661718818126246, 0.913209064863622, 0.435183210065588, 0.205517880385742, 0.645052736159414, 0.833247889531776, 0.812012037960812, 0.395130272954702, 0.00754968752153218, 0.498955159448087, 0.442329543177038, 0.844619557959959, 0.052673046477139, 0.351328540593386, 0.398362676147372, 0.0448020747862756, 0.129515876295045, 0.00492870714515448, 0.741074091754854, 0.266646603122354, 0.19224355975166, 0.40585910435766, 0.918131632264704, 0.193383082281798, 0.220469793770462, 0.0505376176442951, 0.508657294558361, 0.685529666021466, 0.735260733868927, 0.174067105632275, 0.780683270189911, 0.751737065380439, 0.556376125663519, 0.143584269797429, 0.470079771708697, 0.279939191881567, 0.551947300555184, 0.0518394273240119, 0.498796973144636, 0.839127142215148, 0.766924805939198, 0.0283681745640934, 0.460124063305557, 0.44217884587124, 0.825177734950557, 0.691918644588441, 0.906976126600057, 0.0454617822542787, 0.344118683598936, 0.208802431821823, 0.222914564423263, 0.033836194081232, 0.363737218081951, 0.93958752322942, 0.23075060499832, 0.403448767960072, 0.598355867899954, 0.0959147606045008, 0.720115568488836, 0.639313710154966, 0.576604781672359, 0.589908510213718, 0.486228967551142, 0.917349274503067, 0.260209997650236, 0.87815621914342, 0.687049875035882, 0.111724427901208, 0.91566232778132, 0.550102998502553, 0.773510770406574, 0.0649064811877906, 0.943675064248964, 0.924500534310937, 0.0464232752565295, 0.60166640393436, 0.0295400398317724, 0.214608073234558, 0.976416750112548, 0.747174070449546, 0.994687035679817, 0.290824100840837, 0.269208600046113, 0.165512582752854, 0.301956043113023, 0.808866461738944, 0.671051616547629, 0.885772036621347, 0.879311494762078, 0.647320938529447, 0.609296415466815, 0.271016946295276, 0.770962751936167, 0.310281675308943, 0.351729959016666, 0.612242012750357, 0.405529787298292, 0.808681451017037, 0.795421274844557, 0.341964315157384, 0.163680319907144, 0.234607592457905, 0.663601482985541, 0.926942577352747, 0.998256389284506, 0.0756043568253517, 0.572895930381492, 0.909713978413492, 0.419239615788683, 0.716044077882543, 0.00655528693459928, 0.930855592014268, 0.603513882961124, 0.355077959131449, 0.261745644267648, 0.931461959611624, 0.0501294643618166, 0.650053259916604, 0.869903697166592, 0.192219950724393, 0.150438750162721, 0.676744322059676, 0.403591732960194, 0.02808144595474, 0.621188972145319, 0.769809146178886, 0.0660707503557205, 0.972199337324128, 0.279904899653047, 0.876650664722547, 0.385415499797091, 0.521312667522579, 0.27402233122848, 0.985225482843816, 0.478097953600809, 0.564344919519499, 0.925348821794614, 0.247349966550246, 0.235351395793259, 0.868373451754451, 0.85041850595735, 0.592782240128145, 0.865250361850485, 0.496015531709418, 0.698475943179801, 0.184400002239272, 0.998622838407755, 0.165523494593799, 0.732628907542676, 0.143808965571225, 0.174250706331804, 0.965758008649573, 0.793458547908813, 0.0217456389218569, 0.296480822609738, 0.592390775214881, 0.752531420672312, 0.899064305005595, 0.168951421976089, 0.901430058991536, 0.984169271541759, 0.439587800996378, 0.661224898183718, 0.122400605585426, 0.807587762363255, 0.126734163844958, 0.462355116615072, 0.18072821199894, 0.791764266090468, 0.469082997646183, 0.712377787800506, 0.336154757766053, 0.243532264605165, 0.202827049186453, 0.890445961616933, 0.864282209426165, 0.460366809740663, 0.328980821417645, 0.811451097717509, 0.188449112232774, 0.885572424158454, 0.8123674322851, 0.514471093658358, 0.581873727962375, 0.872610488208011, 0.210889032809064, 0.105763354571536, 0.518192722927779, 0.43855158961378, 0.0763315809890628, 0.73340192460455, 0.244574127951637, 0.648866307921708, 0.805195142747834, 0.120356067782268, 0.000693708658218384, 0.973582661245018, 0.544253330444917, 0.472226880723611, 0.660565584199503, 0.546413337578997, 0.468123272061348, 0.0364120420999825), actual = c(1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"