Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>       0.00000000       1.00000000       0.84844672       0.09165869 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>       0.50000000       0.49970701       0.49970701       0.52908296 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.144039672566578, 0.613669121172279, 0.624288582941517, 0.11191529687494, 0.220457135234028, 0.973046100698411, 0.753227037843317, 0.506340209394693, 0.439076449722052, 0.856559875188395, 0.134448774391785, 0.821774974698201, 0.411390929948539, 0.241267479723319, 0.253605664940551, 0.0349076853599399, 0.521140028955415, 0.333113159285858, 0.604688743362203, 0.353768551955, 0.582349933683872, 0.555068864719942, 0.987752512097359, 0.56411433708854, 0.935158327454701, 0.31467792368494, 0.744756134226918, 0.518960541114211, 0.389114073943347, 0.783582120668143, 0.506184233818203, 0.473830080823973, 0.971167595824227, 0.683762103319168, 0.197615785291418, 0.708026567706838, 0.0746790929697454, 0.0335620238911361, 0.287072388455272, 0.286877307342365, 0.864793370710686, 0.958741792477667, 0.860765728633851, 0.0781299914233387, 0.718409519875422, 0.740411043632776, 0.402438881341368, 0.814190127654001, 0.417044293601066, 0.499726733425632, 0.86621933686547, 0.444729312788695, 0.0742798438295722, 0.736167832976207, 0.479825718095526, 0.294582760659978, 0.23513535130769, 0.344915582565591, 0.403534267097712, 0.580894629471004, 0.716414547292516, 0.966931835049763, 0.390682952711359, 0.977314424235374, 0.21722800261341, 0.529500970849767, 0.363107551587746, 0.194203173043206, 0.0358776410575956, 0.771036798832938, 0.0782874228898436, 0.0964239579625428, 0.106558889616281, 0.632914542220533, 0.976982280611992, 0.464871136238798, 0.755215191980824, 0.401439550332725, 0.656555525725707, 0.511922389268875, 0.533268322469667, 0.685424306895584, 0.546597338048741, 0.366963770939037, 0.0692957809660584, 0.0694021061062813, 0.959393459372222, 0.278331292560324, 0.419219818897545, 0.684174714609981, 0.72517478489317, 0.232152584241703, 0.662193357245997, 0.492001302307472, 0.419741497375071, 0.533503945218399, 0.840686743147671, 0.545632920227945, 0.76676369109191, 0.902825680095702, 0.528071528533474, 0.971221978310496, 0.628550425404683, 0.89477894641459, 0.894108995096758, 0.743686643661931, 0.867837482597679, 0.959174663526937, 0.114595423918217, 0.338030209532008, 0.361259727971628, 0.530024404171854, 0.595604492584243, 0.823233642149717, 0.52585224667564, 0.190786000806838, 0.208378846757114, 0.702040351461619, 0.944702679757029, 0.846921145683154, 0.513763764174655, 0.860195867717266, 0.2382159284316, 0.683466728311032, 0.344893595436588, 0.942636244697496, 0.24523679446429, 0.574798560934141, 0.547038999618962, 0.788151710294187, 0.0252981085795909, 0.34461998171173, 0.56018771789968, 0.339417518116534, 0.334978304570541, 0.707751878071576, 0.700677522923797, 0.358462759526446, 0.465648202458397, 0.732414356898516, 0.748759269015864, 0.713577802060172, 0.113412032369524, 0.694245976395905, 0.724070374621078, 0.00296184467151761, 0.973146881209686, 0.071898250374943, 0.33703704085201, 0.905730479164049, 0.954799278872088, 0.07738604885526, 0.325118422275409, 0.574951685266569, 0.576971460832283, 0.203716950025409, 0.69737595343031, 0.0244010502938181, 0.262303902534768, 0.0457508924882859, 0.783252457855269, 0.753333020722494, 0.137434981996194, 0.682862998452038, 0.700672251172364, 0.550906618358567, 0.674568888964131, 0.663163107121363, 0.354887923924252, 0.540162288583815, 0.00987406563945115, 0.872467832872644, 0.852948027895764, 0.822903682244942, 0.0476902492810041, 0.720325542381033, 0.539433436235413, 0.410937210079283, 0.74309351737611, 0.247146976413205, 0.553700697142631, 0.945202605798841, 0.723503472981974, 0.706631345208734, 0.949151573702693, 0.784334107302129, 0.0772996891755611, 0.834376045968384, 0.364271481055766, 0.550170068629086, 0.399577671196312, 0.360443007666618, 0.692360344110057, 0.609104360453784, 0.0108882512431592, 0.283949078526348, 0.642162131145597, 0.514864305732772, 0.927757637575269, 0.598625776357949, 0.426852643489838, 0.0538318892940879, 0.496436961228028, 0.729878871003166, 0.883053159574047, 0.661945810541511, 0.110488620353863, 0.0561778042465448, 0.474133721087128, 0.921882627997547, 0.263527267845348, 0.147902643540874, 0.0886654586065561, 0.465323623968288, 0.142204307951033, 0.688187266467139, 0.187793772434816, 0.70316408528015, 0.0658497384283692, 0.176334175979719, 0.30404031672515, 0.460504232672974, 0.654103408800438, 0.120641421061009, 0.0405888692475855, 0.0601612683385611, 0.724116646219045, 0.761400197865441, 0.950904122088104, 0.942721745697781, 0.00145406275987625, 0.377894497243688, 0.869557506404817, 0.878274018876255, 0.615633453242481, 0.169068305753171, 0.262801566626877, 0.212237624684349, 0.957264469470829, 0.928530384320766, 0.786653726594523, 0.912404535105452, 0.829760134918615, 0.945759498281404, 0.311346494127065, 0.269350923830643, 0.046438573859632, 0.308208910515532, 0.518805258674547, 0.536110632121563, 0.268191163893789, 0.666582345729694, 0.00672330474480987, 0.325758995721117, 0.132360990624875, 0.880606898339465, 0.0933878105133772, 0.465583264827728, 0.696867539314553, 0.102451792452484, 0.0766890866216272, 0.999329232145101, 0.135361668653786, 0.594780832994729, 0.340898838127032, 0.368834625696763, 0.469892095774412, 0.0150158694013953, 0.537307929713279, 0.757354479050264, 0.14020795491524, 0.953316786093637, 0.673381512984633, 0.778653933666646, 0.659951830282807, 0.396538942586631, 0.329877682495862, 0.276895737973973, 0.0175058995373547, 0.0972456133458763, 0.0892263192217797, 0.524188976502046, 0.164648174773902, 0.286534927552566, 0.446694596204907, 0.339802559232339, 0.78627243777737, 0.316004209220409, 0.924057898577303, 0.362780599156395, 0.0109606338664889, 0.231635872973129, 0.701527556171641, 0.10104026272893, 0.134914986789227, 0.925201548729092, 0.495467750122771, 0.329393688822165, 0.0337752832565457, 0.648437405237928, 0.773704539053142, 0.588322196388617, 0.151560833677649, 0.483122356235981, 0.956544417887926, 0.747741980478168, 0.738174227531999, 0.513589451555163, 0.881270254729316, 0.772324508056045, 0.0105631782207638, 0.0308633751701564, 0.173228456405923, 0.769167352700606, 0.428895354270935, 0.677924582269043, 0.176174530759454, 0.301356582902372, 0.101225035730749, 0.938772078603506, 0.69542461889796, 0.903877462027594, 0.88876839238219, 0.00480029638856649, 0.590751823270693, 0.794387496542186, 0.863000887446105, 0.924569454975426, 0.502584086265415, 0.0857923137955368, 0.378657665802166, 0.88618604815565, 0.412589317187667, 0.138751789228991, 0.200977547327057, 0.683253417722881, 0.859818096039817, 0.165570836979896, 0.401805589906871, 0.56857249350287, 0.140219706576318, 0.0458825887180865, 0.655823764624074, 0.431765667628497, 0.462621500948444, 0.844040589639917, 0.0148278761189431, 0.712311180541292, 0.224145336542279, 0.231929827481508, 0.483406158629805, 0.944343431852758, 0.0463986571412534, 0.190956750884652, 0.393829197622836, 0.107419720152393, 0.612474425230175, 0.628556053852662, 0.525263319257647, 0.531004561576992, 0.337351416703314, 0.0919608406256884, 0.582159786252305, 0.985075229313225, 0.611983857816085, 0.529139157151803, 0.20898078312166, 0.283745202701539, 0.00601754081435502, 0.39224822469987, 0.492939892457798, 0.561466193525121, 0.815962408436462, 0.484706512186676, 0.763587680878118, 0.669333389028907, 0.635348597308621, 0.104130915598944, 0.215744769666344, 0.206440352834761, 0.859229063149542, 0.581886420026422, 0.368190831737593, 0.759965588571504, 0.365993115585297, 0.216298630693927, 0.716315794503316, 0.843955019721761, 0.52500552078709, 0.142320740269497, 0.876204760046676, 0.469372190302238, 0.67529811989516, 0.694941648514941, 0.578329639974982, 0.455307592637837, 0.0768696253653616, 0.723147934069857, 0.895730241201818, 0.839444052660838, 0.57055482105352, 0.90469996444881, 0.620457446668297, 0.462095508584753, 0.718179132556543, 0.432626339141279, 0.777418504934758, 0.533903280738741, 0.833532558521256, 0.551908856723458, 0.360356627497822, 0.164378222078085, 0.795908447587863, 0.196200784761459, 0.151571024907753, 0.363040999742225, 0.772428449243307, 0.869226374197751, 0.00158444442786276, 0.696858291048557, 0.863212786382064, 0.471541429171339, 0.417503151809797, 0.373773680534214, 0.578653111355379, 0.507855211850256, 0.975908728083596, 0.812631881330162, 0.352269901195541, 0.88272702600807, 0.367099347291514, 0.202375908847898, 0.973019198514521, 0.894320563413203, 0.458581350976601, 0.95374478562735, 0.382584956940264, 0.0673543147277087, 0.521173457149416, 0.573382504750043, 0.134822340449318, 0.0108223867136985, 0.105640496592969, 0.795565444743261, 0.0846974616870284, 0.834458673838526, 0.625586106209084, 0.31407167064026, 0.41144735715352, 0.889414545381442, 0.0974650590214878, 0.126917210640386, 0.259940049145371, 0.185249323025346, 0.157710957108065, 0.829954255139455, 0.546405762201175, 0.213591802632436, 0.217811622424051, 0.129533152794465, 0.494964737212285, 0.594774415250868, 0.934531960170716, 0.684588379226625, 0.293406146811321, 0.822796712862328, 0.761318107368425, 0.71794524602592, 0.796536715701222, 0.953530048020184, 0.568902267608792, 0.868626798735932, 0.296788591658697, 0.246920753503218, 0.929076366825029, 0.164102592505515, 0.567049275385216, 0.680673085618764, 0.82729378156364, 0.81189577979967, 0.333562512416393, 0.246905186213553, 0.753231465118006, 0.998819725122303, 0.313069344265386, 0.476136290933937, 0.670463241403922, 0.870888233417645, 0.0888051164802164, 0.441577374003828, 0.0244119544513524, 0.680286390474066, 0.328951342497021, 0.235788161400706, 0.293127664364874, 0.617339604068547, 0.00101956166327, 0.41305447020568, 0.85411696205847, 0.306490790564567, 0.818154239561409, 0.313030657824129, 0.337232521967962, 0.779913278995082, 0.912513495888561, 0.0638603279367089, 0.440653823083267, 0.132451515179127, 0.511364851845428, 0.272451004944742, 0.786936650518328, 0.426270777825266, 0.0684087756089866, 0.713139137718827, 0.477403852390125, 0.820172807667404, 0.969596635550261, 0.143716417485848, 0.214448645478114, 0.695679326076061, 0.540858148597181, 0.666880280012265, 0.191308516310528, 0.35397982224822, 0.301166173070669, 0.749426661990583, 0.239626115886495, 0.165504475822672, 0.294300171080977, 0.518033294705674, 0.370356736239046, 0.569432259304449, 0.741376308491454, 0.517953027971089, 0.0744220630731434, 0.63489884021692, 0.935594990616664, 0.463939630426466, 0.557318741921335, 0.916340246796608, 0.939082400640473, 0.596550222951919, 0.962114635854959, 0.599576885579154, 0.775327685289085, 0.868380647618324, 0.329640615498647, 0.772836955962703, 0.895827357424423, 0.912384810857475, 0.610407608328387, 0.0510533091146499, 0.427227421663702, 0.991007357137278, 0.139033521525562, 0.249894513050094, 0.955098991980776, 0.897055509267375, 0.928357625612989, 0.204670087900013, 0.514203984523192, 0.79356929496862, 0.75685795256868, 0.256839857669547, 0.672547212336212, 0.24600370391272, 0.439761290093884, 0.0143334800377488, 0.500242556445301, 0.883894692407921, 0.751839926233515, 0.417407818837091, 0.726770787267014, 0.810050310567021, 0.199586902279407, 0.632087329169735, 0.897678023902699, 0.0603609315585345, 0.578940564999357, 0.423767135711387, 0.694434651872143, 0.52865231060423, 0.985161823919043, 0.417274974286556, 0.815827576210722, 0.84193166391924, 0.0494173623155802, 0.437214696314186, 0.931994535494596, 0.324812528677285, 0.524310767883435, 0.322829894721508, 0.461071512196213, 0.00369993969798088, 0.649906070670113, 0.691750325029716, 0.70641786698252, 0.691193582257256, 0.887078782310709, 0.166906857164577, 0.159108242718503, 0.106559772742912, 0.694289984414354, 0.0538284585345536, 0.718506376957521, 0.745952874654904, 0.460800965316594, 0.321286326274276, 0.0712068420834839, 0.727678576949984, 0.8019593663048, 0.831510509829968, 0.955833096755669, 0.129996488336474, 0.243417088873684, 0.37241409602575, 0.46452716505155, 0.51449243654497, 0.672422607429326, 0.0948971170000732, 0.766233611153439, 0.317596154985949, 0.956457660300657, 0.196337935980409, 0.860006037168205, 0.133996023796499, 0.274322113953531, 0.869783245027065, 0.198568779742345, 0.0849951410200447, 0.671488430351019, 0.246047920314595, 0.514637721702456, 0.895670736441389, 0.0619329113978893, 0.086638854816556, 0.716360846301541, 0.0991386577952653, 0.558444929774851, 0.746328161330894, 0.674738885136321, 0.958139268215746, 0.799582895589992, 0.641341933514923, 0.0283260063733906, 0.239446111023426, 0.541584054939449, 0.65571408579126, 0.0459172322880477, 0.454969074809924, 0.0123579166829586, 0.505356875946745, 0.201088362373412, 0.415291043231264, 0.112909303512424, 0.135792572516948, 0.830927464645356, 0.401620447868481, 0.222136044641957, 0.690145128173754, 0.286445861216635, 0.627472535474226, 0.537215770222247, 0.125983784208074, 0.390955350361764, 0.0506944661028683, 0.379819707944989, 0.882926360471174, 0.00471600890159607, 0.860964928520843, 0.855853959685192, 0.242907150881365, 0.0627115718089044, 0.634772337740287, 0.680152382934466, 0.368597684893757, 0.534827319439501, 0.74399034306407, 0.0385964778251946, 0.236663345480338, 0.327236599987373, 0.491168982116506, 0.137410139897838, 0.35401091305539, 0.395414187340066, 0.503736660582945, 0.462262224173173, 0.165497603127733, 0.810336244758219, 0.955258599715307, 0.506369445705786, 0.190275832079351, 0.343108082422987, 0.720810250611976, 0.971272794064134, 0.388888476416469, 0.215748841641471, 0.0774334846064448, 0.138847977621481, 0.129497312242165, 0.10459807375446, 0.29156066919677, 0.999116492457688, 0.0579009279608727, 0.79886677255854, 0.695391758112237, 0.0172360811848193, 0.193209799937904, 0.245410007657483, 0.398050197632983, 0.981836933409795, 0.123258127132431, 0.00356292654760182, 0.693267113761976, 0.596199422609061, 0.599216291215271, 0.902748229214922, 0.0453128830995411, 0.960555278463289, 0.603947666706517, 0.233693620888516, 0.616569031495601, 0.0374582014046609, 0.775734483497217, 0.0237848819233477, 0.0690788021311164, 0.114636884536594, 0.506658277474344, 0.484940143302083, 0.904131496790797, 0.683012987952679, 0.843339734710753, 0.771963344421238, 0.640795953338966, 0.642663044156507, 0.0516437583137304, 0.753756404155865, 0.531879364047199, 0.188189373118803, 0.00730417971499264, 0.943239664658904, 0.0386657842900604, 0.418216080637649, 0.845660656690598, 0.708321539917961, 0.912665021140128, 0.103997653815895, 0.631605500355363, 0.0185896337497979, 0.205608717165887, 0.584658972686157, 0.52703052549623, 0.0572019917890429, 0.775650934549049, 0.970434811664745, 0.220872356556356, 0.817704878281802, 0.427704620407894, 0.122533148154616, 0.873029510257766, 0.805906685767695, 0.418421242386103, 0.746381704695523, 0.209859146270901, 0.0570433035027236, 0.167973232688382, 0.356133493361995, 0.66210030997172, 0.756700613070279, 0.42099616676569, 0.510196008253843, 0.0848868717439473, 0.662594574503601, 0.418040738673881, 0.804417387815192, 0.285065884003416, 0.887026653857902, 0.507027132203802, 0.458927977830172, 0.231245383154601, 0.704714153893292, 0.417647915426642, 0.14830243610777, 0.854450513608754, 0.599582791561261, 0.718713412294164, 0.0525090307928622, 0.501232420327142, 0.208642864599824, 0.517900713486597, 0.828533579362556, 0.700495701981708, 0.12358787143603, 0.560073681175709, 0.6362521185074, 0.874952062265947, 0.150905974442139, 0.0480378014035523, 0.614196705631912, 0.135440094862133, 0.945478726178408, 0.984943191055208, 0.432958773570135, 0.668017510091886, 0.2555753714405, 0.464147253427655, 0.706181624438614, 0.143635484622791, 0.825503680622205, 0.680356862721965, 0.201714773895219, 0.041344235651195, 0.188650861848146, 0.201924023451284, 0.832918810425326, 0.414867483312264, 0.866018776549026, 0.321928696939722, 0.294995113508776, 0.217301596188918, 0.0274501238018274, 0.758260981412604, 0.487563236150891, 0.165849547134712, 0.5513465621043, 0.772003963822499, 0.0299919859971851, 0.362627209164202, 0.202369567705318, 0.774008979555219, 0.240090522915125, 0.5179114039056, 0.730545895406976, 0.345578544074669, 0.0508587900549173, 0.708437521476299, 0.022554671857506, 0.110815849155188, 0.288138574222103, 0.0657534948550165, 0.843946750741452, 0.288211140083149, 0.500519710592926, 0.5692019299604, 0.387140802573413, 0.797547992086038, 0.735395129304379, 0.26366892689839, 0.87053114338778, 0.384112891508266, 0.583773680962622, 0.990291710710153, 0.0592899254988879, 0.602765281451866, 0.0660948255099356, 0.0348417060449719, 0.883533477783203, 0.136284393258393, 0.789576141862199, 0.656021822243929, 0.397655365755782, 0.847352719400078, 0.83817314915359, 0.887864619260654, 0.721697421045974, 0.0499630412086844, 0.858266014605761, 0.00428205216303468, 0.00235077855177224, 0.356543825240806, 0.865550618385896, 0.602894330630079, 0.74425885360688, 0.368678318802267, 0.399884512647986, 0.209543384378776, 0.417218876071274, 0.28597856964916, 0.243156734853983, 0.433205739129335, 0.335625171894208, 0.213334979722276, 0.590989446733147, 0.707794961519539, 0.307618694147095, 0.974685405846685, 0.55391617002897, 0.920991711784154, 0.0452467419672757, 0.171084392350167, 0.143864797661081, 0.747446130495518, 0.591274663805962, 0.0735610343981534, 0.0160399423912168, 0.247001843992621, 0.675060304114595, 0.380287664942443, 0.939911832800135, 0.479670448228717, 0.61341945361346, 0.7467242393177, 0.324627273483202, 0.487388548208401, 0.719673213316128, 0.907181113492697, 0.125567328184843, 0.352596992626786, 0.987634236225858, 0.0379177529830486, 0.797993201995268, 0.628221745369956, 0.635833569103852, 0.992018589749932, 0.52661799127236, 0.256990047171712, 0.488345648394898, 0.166350811021402, 0.493463785154745, 0.740212865639478, 0.797845909604803, 0.728491425747052, 0.156667254632339, 0.0927160584833473, 0.34503287053667, 0.472662753658369, 0.346195795806125, 0.160787790082395, 0.374389958800748, 0.445645730243996, 0.329703956376761, 0.701069419970736, 0.336939502507448, 0.571023460011929, 0.388235968537629, 0.298271896317601, 0.321977003244683, 0.536958334036171, 0.317517573479563, 0.830095326760784, 0.251123652327806, 0.644525660201907, 0.843367067165673, 0.931374510750175, 0.0346987417433411, 0.686641260515898, 0.352699898183346, 0.895202931482345, 0.8039075715933, 0.353843624936417, 0.548691518604755, 0.892108526313677, 0.00399380689486861, 0.369220827938989, 0.936902508838102, 0.871668560197577, 0.463783396175131, 0.400219559203833, 0.114522116957232, 0.971234023338184, 0.415496101137251, 0.895385749172419, 0.816924409708008, 0.54473194363527, 0.78288708999753, 0.346991210011765, 0.539219874655828, 0.105479742866009, 0.513623086502776, 0.78103893622756, 0.519905440276489, 0.102073973976076, 0.298695611767471, 0.36354957614094, 0.142480223439634, 0.814149319892749, 0.18078065966256, 0.499035901390016, 0.686277249129489, 0.198801767313853, 0.126456541242078, 0.35783073795028, 0.273052083095536, 0.347146165790036, 0.293109450489283, 0.405578845646232, 0.331617128569633, 0.100425783311948, 0.721822499297559, 0.265394841320813, 0.248632440110669, 0.114254074171185, 0.161751401377842, 0.898484692210332, 0.610132979694754, 0.598084545927122, 0.0138203566893935, 0.827565067913383, 0.241475684102625, 0.256106168264523), actual = c(0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.132895376766101, 0.477476512547582, 0.059902255423367, 0.169797625625506, 0.490083361975849, 0.00371412024833262, 0.918988175224513, 0.0977717973291874, 0.464840816799551, 0.353591403691098, 0.078598384745419, 0.83159643644467, 0.790357250720263, 0.469028699211776, 0.615266794338822, 0.132185543887317, 0.539643231080845, 0.176536263199523, 0.0374130047857761, 0.822150807827711, 0.178331237053499, 0.129674459807575, 0.115154538303614, 0.327001858735457, 0.528500881278887, 0.930116481613368, 0.261883986182511, 0.392909459769726, 0.170027535874397, 0.194943764479831, 0.279129273956642, 0.707810827996582, 0.979777426226065, 0.101717211073264, 0.10098773939535, 0.00532718421891332, 0.757411427097395, 0.575093758059666, 0.0521533761639148, 0.581248052185401, 0.530744135845453, 0.647208457347006, 0.592857132665813, 0.90014988835901, 0.669069042196497, 0.828820359893143, 0.561662561260164, 0.728781769517809, 0.81309203337878, 0.24441606644541, 0.851846446748823, 0.181954777101055, 0.785847295308486, 0.418527435511351, 0.576407702174038, 0.401566450949758, 0.483204147545621, 0.740130927879363, 0.289315189002082, 0.143365976866335, 0.365137342363596, 0.168240656610578, 0.835044054547325, 0.569491151487455, 0.46456079906784, 0.663125678896904, 0.576572339283302, 0.234053156105801, 0.0515597115736455, 0.413029341259971, 0.709467817796394, 0.888189124874771, 0.00706287706270814, 0.929143620654941, 0.269172768108547, 0.13583336956799, 0.918549813330173, 0.694059333298355, 0.355836026137695, 0.760335453785956, 0.0615837203804404, 0.369653193280101, 0.501146899536252, 0.0720295361243188, 0.568984243320301, 0.0935252527706325, 0.839634401025251, 0.933240284211934, 0.434911683900282, 0.638173981569707, 0.473966992460191, 0.321238112170249, 0.740446890005842, 0.251307474449277, 0.733912431634963, 0.966287353076041, 0.946494614239782, 0.0742414575070143, 0.663062336156145, 0.185149904340506, 0.915586290881038, 0.432934253010899, 0.995402371976525, 0.693267480935901, 0.0122461630962789, 0.0758147465530783, 0.288083236664534, 0.598373885499313, 0.577206169022247, 0.209384043933824, 0.310905686346814, 0.559382967883721, 0.0657745501957834, 0.917153090471402, 0.320529777090997, 0.676258405670524, 0.343488065525889, 0.353562659816816, 0.50077926623635, 0.823324533645064, 0.741960755549371, 0.74671356077306, 0.855777549091727, 0.387602647999302, 0.0334628773853183, 0.627800793386996, 0.395040207076818, 0.120094479992986, 0.77828812925145, 0.706602588761598, 0.392229310935363, 0.919256828026846, 0.647816357668489, 0.230020300485194, 0.994742369512096, 0.727581183891743, 0.265446725999936, 0.30373199749738, 0.888587602647021, 0.71157832769677, 0.409823911264539, 0.367800305131823, 0.523930227383971, 0.0145421801134944, 0.145634474931285, 0.509852331597358, 0.317716960329562, 0.235196677502245, 0.368454382522032, 0.961446661036462, 0.642986519960687, 0.863832511706278, 0.790137498406693, 0.204535966739058, 0.989820059156045, 0.0199897352140397, 0.852839241037145, 0.192259298637509, 0.733314174227417, 0.356120524229482, 0.830027702497318, 0.321705949725583, 0.680424868827686, 0.414679416455328, 0.961618794826791, 0.855384654598311, 0.494507229421288, 0.483076258096844, 0.555525881471112, 0.630481368163601, 0.880308701889589, 0.553744660457596, 0.873731575207785, 0.55252421926707, 0.52810938633047, 0.323998456355184, 0.465777497505769, 0.062610847176984, 0.238391446415335, 0.353093389188871, 0.0334766646847129, 0.0146126113831997, 0.6663122640457, 0.533564254408702, 0.913154597394168, 0.277436102740467, 0.303318700985983, 0.799585995497182, 0.953361912397668, 0.555056540528312, 0.0678696611430496, 0.52277419785969, 0.373387491796166, 0.984912208979949, 0.609642442781478, 0.540119593031704, 0.945931805064902, 0.835789227159694, 0.815602926537395, 0.798282983247191, 0.784981446573511, 0.656163224484771, 0.0860797290224582, 0.00813575554639101, 0.677014384884387, 0.0699288945179433, 0.0268036832567304, 0.883434964809567, 0.552656936459243, 0.0174526770133525, 0.642694085370749, 0.851983665954322, 0.439853044925258, 0.761321841273457, 0.0723546240478754, 0.0687776249833405, 0.57057457906194, 0.286438007140532, 0.506341511150822, 0.162366255652159, 0.202849015593529, 0.40010563400574, 0.494254239136353, 0.704135119915009, 0.286490387981758, 0.00970726436935365, 0.0562265580520034, 0.890792581252754, 0.250659621087834, 0.945431693457067, 0.0987876418512315, 0.188437731703743, 0.203903445741162, 0.976084440713748, 0.248362691141665, 0.531094804406166, 0.58592762472108, 0.279550210107118, 0.734969285782427, 0.265670166350901, 0.090541445882991, 0.967257692245767, 0.398773618740961, 0.692537650233135, 0.847121583297849, 0.366309070261195, 0.825354164233431, 0.0949706598185003, 0.1206253929995, 0.711199606535956, 0.587789263576269, 0.397048060083762, 0.364426128100604, 0.920103017939255, 0.986327295424417, 0.75827946444042, 0.157767317956313, 0.316085051978007, 0.00794903980568051, 0.319939270615578, 0.565326240379363, 0.825029682600871, 0.298266685800627, 0.988785328809172, 0.348524548811838, 0.897255938034505, 0.955473097972572, 0.925106544280425, 0.632607870036736, 0.687047762330621, 0.478900810936466, 0.405100244330242, 0.483600061852485, 0.893868983024731, 0.648230133811012, 0.845399838872254, 0.263870908413082, 0.302252114983276, 0.00160839292220771, 0.0294234529137611, 0.875225620809942, 0.675543816294521, 0.830710527487099, 0.546237390255556, 0.98018621490337, 0.429767550434917, 0.0974259178619832, 0.489151389338076, 0.14188992488198, 0.0129755835514516, 0.591625845525414, 0.052026922814548, 0.4796695089899, 0.729670693399385, 0.304155633784831, 0.517328201560304, 0.536828314187005, 0.323116574203596, 0.150914001977071, 0.503451795550063, 0.256198656512424, 0.0686214831657708, 0.693936249008402, 0.860060866456479, 0.0922186030074954, 0.490315719740465, 0.526471060235053, 0.413814040832222, 0.509529631352052, 0.0350425811484456, 0.30761477118358, 0.925118909217417, 0.196795402094722, 0.0217476855032146, 0.994114357978106, 0.261629396583885, 0.238022082485259, 0.767356066033244, 0.773930432042107, 0.278743966016918, 0.854574686381966, 0.301511293277144, 0.467488190624863, 0.336897463770583, 0.0448900831397623, 0.44168419088237, 0.0238074625376612, 0.135443611536175, 0.730605373391882, 0.343877428909764, 0.0485935397446156, 0.452124023577198, 0.00523597397841513, 0.68317260267213, 0.164586088852957, 0.772329724859446, 0.752125042956322, 0.788054569158703, 0.957973246229813, 0.453073123702779, 0.0405188510194421, 0.543103440199047, 0.630656770430505, 0.661205586278811, 0.421062004519626, 0.335977627662942, 0.932569825323299, 0.235135046998039, 0.769037001999095, 0.524282471742481, 0.0394979086704552, 0.0800097316969186, 0.27373573416844, 0.837550509022549, 0.26220301957801, 0.728517887648195, 0.739443870261312, 0.0421873545274138, 0.293067728169262, 0.238654710818082, 0.586943238042295, 0.234340822556987, 0.357809612993151, 0.490855564363301, 0.3514526237268, 0.131165963597596, 0.732046084944159, 0.835015346063301, 0.411349968286231, 0.51556219579652, 0.118773232446983, 0.674024759558961, 0.372906211065128, 0.874912132974714, 0.920861616032198, 0.637077772989869, 0.611070591956377, 0.107571985339746, 0.610805606236681, 0.301180386682972, 0.363740117987618, 0.0672405606601387, 0.707749638939276, 0.240093935048208, 0.528156220680103, 0.625947688473389, 0.939356417395175, 0.311519634211436, 0.824055335484445, 0.00583709543570876, 0.476531055988744, 0.526408854173496, 0.0324232422281057, 0.412090251920745, 0.485610123258084, 0.781020191730931, 0.225392502499744, 0.0440368209965527, 0.896255738567561, 0.98000450944528, 0.0965977993328124, 0.850584610365331, 0.198455039877445, 0.753419807413593, 0.790335008176044, 0.052057109773159, 0.939197283936664, 0.656032118014991, 0.519460908137262, 0.31718837050721, 0.573738494422287, 0.620283094001934, 0.490330734290183, 0.164628370897844, 0.0762577929999679, 0.0126838765572757, 0.313109140144661, 0.255320152267814, 0.697402035584673, 0.438980867387727, 0.315501019824296, 0.26321119396016, 0.138322627637535, 0.703617741819471, 0.143540100427344, 0.760837425012141, 0.50577215035446, 0.107547644060105, 0.596782912733033, 0.641937489854172, 0.335692395456135, 0.925450178794563, 0.450821675593033, 0.630769958952442, 0.436017562868074, 0.924179890891537, 0.915169773856178, 0.359865982085466, 0.586884051794186, 0.0392941397149116, 0.514030181569979, 0.170925264013931, 0.419055788079277, 0.435114042833447, 0.0344292598310858, 0.811740694800392, 0.615346672246233, 0.922393873566762, 0.252114517148584, 0.896785001503304, 0.244133885484189, 0.166704527102411, 0.37643152102828, 0.819013834698126, 0.497864628443494, 0.354214745573699, 0.0439186173025519, 0.376484874868765, 0.341648183763027, 0.91152864927426, 0.122021207585931, 0.253169667208567, 0.950600126525387, 0.290302417241037, 0.87085037608631, 0.332787979394197, 0.89748002262786, 0.968149004969746, 0.211223224177957, 0.743753101211041, 0.391498773358762, 0.911302528111264, 0.41795757249929, 0.702134881401435, 0.569167673354968, 0.982523186597973, 0.210469999816269, 0.352480279514566, 0.175824262434617, 0.447782885283232, 0.586461870931089, 0.568505650851876, 0.733293150784448, 0.505984971532598, 0.592517047189176, 0.946672027697787, 0.948164520552382, 0.414345594355837, 0.176077754469588, 0.340708324452862, 0.953588292468339, 0.235400409437716, 0.0371907211374491, 0.953309438889846, 0.933226815890521, 0.687306377803907, 0.561309913173318, 0.174353886395693, 0.434979132842273, 0.941152474610135, 0.75160798500292, 0.0713624064810574, 0.779287642333657, 0.708562920568511, 0.00066139898262918, 0.424538926687092, 0.671555104665458, 0.931133646285161, 0.00270973611623049, 0.0270667257718742, 0.846196065889671, 0.896170619176701, 0.629245003918186, 0.286154478788376, 0.577649481361732, 0.9312131581828, 0.542835250031203, 0.0620298048015684, 0.826632694574073, 0.923698048107326, 0.957546806195751, 0.874225275823846, 0.372434017714113, 0.164197855163366, 0.815776850096881, 0.795763531699777, 0.760175894480199, 0.996687220875174, 0.141190664842725, 0.309064801083878, 0.120272978208959, 0.41728062578477, 0.595049341907725, 0.652767530176789, 0.886008050758392, 0.894384893355891, 0.76275955257006, 0.283863730728626, 0.928213411010802, 0.622136626392603, 0.642770457314327, 0.754096399294212, 0.772709917044267, 0.0802797614596784, 0.436454811831936, 0.718091362854466, 0.823881810065359, 0.219459492713213, 0.323266240302473, 0.420557581353933, 0.888951919507235, 0.984901146264747, 0.235696595627815, 0.939179682172835, 0.00314700719900429, 0.21335814287886, 0.182200994109735, 0.557347322115675, 0.584256152855232, 0.90669216774404, 0.890340642770752, 0.779554058099166, 0.0252224255818874, 0.404713532654569, 0.118441036203876, 0.30024850810878, 0.813604478724301, 0.338213653070852, 0.930054810363799, 0.691135467961431, 0.748782137176022, 0.844625280471519, 0.824421358760446, 0.762209191918373, 0.519212352810428, 0.956551436567679, 0.828138545854017, 0.821183364605531, 0.336267244070768, 0.737917999038473, 0.585297044832259, 0.363547130255029, 0.925292066764086, 0.155849069589749, 0.881874043028802, 0.387792913941666, 0.87255664402619, 0.633868817007169, 0.959570645354688, 0.324440621305257, 0.315119105158374, 0.371485357638448, 0.0439117602072656, 0.66380436043255, 0.146818164037541, 0.640434678876773, 0.369440509239212, 0.134073972702026, 0.557368928333744, 0.116806974401698, 0.762536271242425, 0.710600134218112, 0.701773463748395, 0.0470148471649736, 0.00219397316686809, 0.0382252372801304, 0.258992361836135, 0.396550464443862, 0.584925242466852, 0.99391324701719, 0.306817496893927, 0.846299274591729, 0.315196390030906, 0.960606785258278, 0.644307404523715, 0.504983358550817, 0.926096971612424, 0.756949890172109, 0.0437199743464589, 0.926492534577847, 0.335718359565362, 0.241235029418021, 0.717149118660018, 0.574373374693096, 0.572051303926855, 0.755190461408347, 0.374316918663681, 0.076504600699991, 0.509601733647287, 0.440153653966263, 0.570065778214484, 0.345731104491279, 0.479650121647865, 0.879563324619085, 0.429619423579425, 0.959993045777082, 0.263917655218393, 0.275811194442213, 0.377043972723186, 0.26620728103444, 0.832679231418297, 0.765403408557177, 0.555908944224939, 0.404701823368669, 0.505336361005902, 0.531081061577424, 0.762437170371413, 0.215432338882238, 0.422909962479025, 0.963724337285385, 0.285005928948522, 0.109490859089419, 0.0983471607323736, 0.787851381581277, 0.247137531870976, 0.16709996922873, 0.493554586777464, 0.0470789263490587, 0.151493344223127, 0.836508395383134, 0.837009929353371, 0.495232975808904, 0.0931739141233265, 0.489148385589942, 0.716686515137553, 0.628478800877929, 0.738512263866141, 0.930329907219857, 0.832848949357867, 0.454330952139571, 0.563603625865653, 0.838986217742786, 0.479139665374532, 0.929559198440984, 0.651623080018908, 0.294834615429863, 0.921961692161858, 0.0728570155333728, 0.376913082553074, 0.144114735536277, 0.0254024318419397, 0.346325270831585, 0.867707282770425, 0.16595584875904, 0.492190109798685, 0.152762157376856, 0.00677722785621881, 0.251238957280293, 0.870039579458535, 0.712122620781884, 0.193907914916053, 0.500719334930182, 0.713661356596276, 0.461122353095561, 0.714514670195058, 0.670963307842612, 0.292017377912998, 0.515992566710338, 0.367274668999016, 0.613255891250446, 0.159461668459699, 0.389579234877601, 0.411879019811749, 0.828297472791746, 0.0700544184073806, 0.481550280936062, 0.604644622420892, 0.955802435986698, 0.2971368834842, 0.622070904588327, 0.447378716664389, 0.952654638327658, 0.395488693146035, 0.847687616245821, 0.0180101983714849, 0.147179435705766, 0.0706359599716961, 0.808410150697455, 0.562751258024946, 0.967451574746519, 0.597860830370337, 0.618958234786987, 0.188991084462032, 0.510440622922033, 0.517870037583634, 0.136121252784505, 0.835699877003208, 0.304948552045971, 0.373880349798128, 0.935510319890454, 0.244467723416165, 0.456045087892562, 0.483362338272855, 0.292083075270057, 0.0525566532742232, 0.125252473400906, 0.294876065803692, 0.736752021126449, 0.164731089491397, 0.404887630837038, 0.5689326019492, 0.902588490396738, 0.129959812387824, 0.662631270010024, 0.0618724131491035, 0.997178834863007, 0.202132304431871, 0.69834892754443, 0.707378054736182, 0.703871917678043, 0.455543686635792, 0.618414486991242, 0.01033697091043, 0.443439709721133, 0.111890103202313, 0.209612988634035, 0.617105374578387, 0.211145137669519, 0.230660012224689, 0.268727066228166, 0.96676186285913, 0.195551913464442, 0.286930963397026, 0.622726097237319, 0.48653325275518, 0.933818847173825, 0.42154752695933, 0.231028276262805, 0.758675882825628, 0.414485111366957, 0.165591949597001, 0.735584806883708, 0.336715263314545, 0.213569119572639, 0.313778129173443, 0.050946953939274, 0.366317733656615, 0.642171261832118, 0.678557025035843, 0.593634979566559, 0.245991113130003, 0.146070537855849, 0.165543321287259, 0.0600797752849758, 0.538760148221627, 0.460078701609746, 0.925632356666028, 0.159748933510855, 0.855830244952813, 0.639286500634626, 0.450927020749077, 0.0467113656923175, 0.224506799830124, 0.293179791420698, 0.172102389624342, 0.560567129170522, 0.288348197704181, 0.622450377093628, 0.961208200547844, 0.0860317379701883, 0.812762923305854, 0.876966227777302, 0.791862804209813, 0.83156619919464, 0.894733243621886, 0.11960005084984, 0.435514778131619, 0.178079256089404, 0.834573554107919, 0.382997943321243, 0.762280304916203, 0.14984127599746, 0.0184737709350884, 0.165900935186073, 0.640416892478243, 0.110227327328175, 0.657954325899482, 0.0797393491957337, 0.115564137231559, 0.573019393254071, 0.916175810154527, 0.0434920799452811, 0.686780584743246, 0.79990399396047, 0.975067575927824, 0.132300801575184, 0.829531392082572, 0.594747322611511, 0.21064349100925, 0.786145105957985, 0.17014904669486, 0.786669535562396, 0.726031033322215, 0.665715240407735, 0.315185991115868, 0.43670841306448, 0.413913693744689, 0.632095588603988, 0.614902390865609, 0.580783640034497, 0.722146839136258, 0.221383044496179, 0.241342703811824, 0.603732307674363, 0.569069611374289, 0.596935826353729, 0.884789795381948, 0.322270189877599, 0.417931071715429, 0.135595254832879, 0.360740927513689, 0.567713782656938, 0.720420280238613, 0.305811600526795, 0.0102743057068437, 0.229347242275253, 0.477123322896659, 0.914549633394927, 0.124903105664998, 0.379537393571809, 0.944216226460412, 0.547113589011133, 0.929039634298533, 0.575528644956648, 0.144098582677543, 0.0292631986085325, 0.203741593984887, 0.740455868653953, 0.521683828206733, 0.852758812252432, 0.951384562533349, 0.284508175216615, 0.77962327003479, 0.644155462272465, 0.915360974846408, 0.278800028143451, 0.148181509459391, 0.0671934178099036, 0.425154332537204, 0.954225867055357, 0.789759839884937, 0.665774087421596, 0.934037634171546, 0.775847896700725, 0.879841248039156, 0.568640654208139, 0.433438277803361, 0.3662611967884, 0.975777817890048, 0.54621307272464, 0.437382719712332, 0.433817411540076, 0.16547211096622, 0.574808474630117, 0.882479544263333, 0.76707290369086, 0.146966256666929, 0.299191488185897, 0.342204198008403, 0.354856570018455, 0.540199336595833, 0.357325280318037, 0.411920572863892, 0.571249743690714, 0.79759802413173, 0.97587615204975, 0.0287546529434621, 0.922757372260094, 0.713919518748298, 0.572645421605557, 0.158613786799833, 0.972509491024539, 0.219808510970324, 0.331953137880191, 0.171471860259771, 0.765165971824899, 0.305736049776897, 0.00291424570605159, 0.36763580259867, 0.962241673143581, 0.307372326729819, 0.398160222452134, 0.947384230559692, 0.0716918939724565, 0.847778661409393, 0.8541629165411, 0.859153567813337, 0.0246634462382644, 0.712380613666028, 0.220309318276122, 0.258726308587939, 0.465677394066006, 0.296864484902471, 0.191560846520588, 0.260788935935125, 0.225439592497423, 0.181570773245767, 0.84490098268725, 0.998168280348182, 0.94796809181571, 0.686344052432105, 0.466444450896233, 0.465434846002609, 0.164148051990196, 0.330234542489052, 0.292886989889666, 0.983768513891846, 0.103541167685762, 0.520056411158293, 0.586714971810579, 0.439113745000213, 0.663368871202692, 0.59937016479671, 0.611185719957575, 0.478211157256737, 0.078302109381184, 0.763518304098397, 0.111885674763471, 0.474863678216934, 0.782854890450835, 0.511817098129541, 0.608519215369597, 0.367619882337749, 0.131213694810867, 0.281919381581247, 0.73453693697229, 0.326556215761229, 0.0952019207179546, 0.077596768271178, 0.356709165498614, 0.58215647446923, 0.0778950150124729, 0.565017188200727, 0.486566238803789, 0.651011144276708, 0.665704993996769, 0.763566675130278, 0.34238251298666, 0.323377559194341, 0.841390566201881, 0.162844207370654, 0.436047242255881, 0.842100014444441, 0.962136207148433, 0.65430070576258, 0.698782221414149, 0.876100418623537, 0.507261068560183, 0.392397630494088, 0.826847148593515, 0.14777539880015, 0.544812010135502, 0.497957991901785, 0.401394319487736, 0.00165876490063965, 0.920201413799077, 0.083688793471083, 0.296581350965425, 0.409427153179422, 0.450203434564173), actual = c(0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"