Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.9807083        0.3363466 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4901646        0.4901646        0.5016688 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.729943877318874, 0.0912213926203549, 0.0412983726710081, 0.662245160434395, 0.131936431163922, 0.334957273676991, 0.087279753992334, 0.258716624230146, 0.923846732825041, 0.663373047485948, 0.759592884685844, 0.356003295397386, 0.118896605679765, 0.467090804828331, 0.0502746184356511, 0.0721123924013227, 0.793964197393507, 0.768849480198696, 0.827557547483593, 0.403152552433312, 0.483931169845164, 0.220244362251833, 0.614927234128118, 0.292480744654313, 0.494888101005927, 0.947584851179272, 0.822143581928685, 0.921268369071186, 0.731067575747147, 0.871227460447699, 0.335717831505463, 0.388506038114429, 0.547568470705301, 0.057096878066659, 0.823649043682963, 0.910882511409, 0.792997489916161, 0.968638171441853, 0.340803833911195, 0.131309612421319, 0.156503323698416, 0.834776943083853, 0.773169975727797, 0.714057328412309, 0.774543969659135, 0.436723206425086, 0.516709205927327, 0.178902042098343, 0.0645514151547104, 0.373506853589788, 0.223358765710145, 0.875618858495727, 0.514087573159486, 0.769331646151841, 0.540906427660957, 0.28583020856604, 0.68953287997283, 0.95690827537328, 0.940516235306859, 0.474317004904151, 0.440277664922178, 0.0312460551504046, 0.88987340638414, 0.819914824562147, 0.325191932963207, 0.0081455884501338, 0.274761471897364, 0.4977935932111, 0.0661221938207746, 0.80514096072875, 0.350251027615741, 0.628730490803719, 0.326191517291591, 0.865344058489427, 0.957464705221355, 0.831015908159316, 0.477507814066485, 0.089956077048555, 0.0374555967282504, 0.10226486925967, 0.452710449462757, 0.0786109962500632, 0.0185866903048009, 0.70958129549399, 0.48376118321903, 0.874394652666524, 0.358528052689508, 0.526955062989146, 0.558205081848428, 0.816566570894793, 0.884091293439269, 0.841428007464856, 0.524449223419651, 0.415309024741873, 0.590910393511876, 0.63683128124103, 0.887176659889519, 0.357449883595109, 0.823576316703111, 0.410933347651735, 0.406730817630887, 0.699041903484613, 0.686973164090887, 0.66251238877885, 0.133582171751186, 0.613448948599398, 0.593507481040433, 0.928591539617628, 0.0058460314758122, 0.544494179775938, 0.907840710831806, 0.383953247219324, 0.239801363553852, 0.807615735568106, 0.708699407055974, 0.0431483131833375, 0.266870772000402, 0.28216556366533, 0.348065526690334, 0.833009720081463, 0.699366081738845, 0.298517854651436, 0.260610508965328, 0.804184914100915, 0.451448663603514, 0.93042805348523, 0.93263780279085, 0.888007705565542, 0.760021964320913, 0.721978509565815, 0.309261090354994, 0.933275219518691, 0.13100558356382, 0.872048617107794, 0.263710202183574, 0.248583140783012, 0.309024798450992, 0.192612382816151, 0.853910774225369, 0.359425810398534, 0.828022309811786, 0.922997067682445, 0.818126938072965, 0.791407580720261, 0.871939952019602, 0.0810509966686368, 0.0937674366869032, 0.735264763236046, 0.279269912047312, 0.515930503839627, 0.842730170115829, 0.026034970767796, 0.681992511963472, 0.278832227922976, 0.586917988257483, 0.277915204875171, 0.593385014217347, 0.854428838007152, 0.378643008181825, 0.673118711216375, 0.574275051476434, 0.975733746541664, 0.0489838279318064, 0.341792945517227, 0.902552098035812, 0.478054974926636, 0.339455212233588, 0.317746538668871, 0.894096963107586, 0.739767511142418, 0.627952649723738, 0.173529982101172, 0.985970314824954, 0.344435945618898, 0.456040908815339, 0.670983102405444, 0.345084573142231, 0.580710038309917, 0.0551520872395486, 0.834051830461249, 0.603322999086231, 0.683277802076191, 0.605333314044401, 0.7629840944428, 0.703578849090263, 0.017290621297434, 0.163825649069622, 0.756096896482632, 0.16792144626379, 0.428503107046708, 0.46874066023156, 0.895745971472934, 0.919319153064862, 0.425165974535048, 0.340720518957824, 0.256176604423672, 0.822749515762553, 0.476531339110807, 0.231830017175525, 0.92229569889605, 0.236932245083153, 0.656165470834821, 0.653948013670743, 0.432980647776276, 0.755098183406517, 0.637898684944957, 0.717604040866718, 0.797202652320266, 0.112173511413857, 0.652673840522766, 0.612831234466285, 0.449097885517403, 0.691255773883313, 0.724586670752615, 0.806238651275635, 0.220897477585822, 0.118107218528166, 0.655314175179228, 0.945152081549168, 0.972977835685015, 0.760019276756793, 0.282394274137914, 0.141140237217769, 0.207578744040802, 0.936997918877751, 0.614734001690522, 0.627384854713455, 0.837471603183076, 0.815190366702154, 0.138994720764458, 0.439392425585538, 0.882621780503541, 0.0274554302450269, 0.0483374730683863, 0.668049880303442, 0.135123582556844, 0.548124213470146, 0.0889352792873979, 0.880251015303656, 0.989671017276123, 0.179695828119293, 0.0354327077511698, 0.131012985948473, 0.804997344035655, 0.941835093777627, 0.532896847231314, 0.562483035493642, 0.447908739326522, 0.306541552068666, 0.366348577896133, 0.906324573326856, 0.627045328728855, 0.203760082367808, 0.896922980900854, 0.8600017381832, 0.276548255933449, 0.380473995115608, 0.122259631054476, 0.0279264554847032, 0.145798199577257, 0.970873595681041, 0.137415086152032, 0.864961205748841, 0.717625184683129, 0.365627040853724, 0.94494249811396, 0.590620072558522, 0.394136348273605, 0.22314910008572, 0.564282836392522, 0.725334692047909, 0.0316620215307921, 0.677459384780377, 0.288119355682284, 0.541311553679407, 0.027051005512476, 0.179688079981133, 0.359630377264693, 0.279455228475854, 0.993062184890732, 0.689215303165838, 0.0342906923033297, 0.200261858059093, 0.216389905195683, 0.571097753476351, 0.987427865155041, 0.520910128718242, 0.147893723798916, 0.20996985048987, 0.656741806073114, 0.787085715681314, 0.311127669876441, 0.243363270070404, 0.588439275044948, 0.673781028017402, 0.834215595852584, 0.0630836479831487, 0.624867937294766, 0.0959801108110696, 0.255427770782262, 0.505693870130926, 0.0119248176924884, 0.688092019641772, 0.108445686521009, 0.0580817447043955, 0.905340537196025, 0.755682141752914, 0.837414648383856, 0.936813862295821, 0.314197471830994, 0.113071557832882, 0.378730878699571, 0.177750116214156, 0.308546301210299, 0.726760622113943, 0.244389197789133, 0.525657686637715, 0.73270396143198, 0.863458218285814, 0.698656592983752, 0.623567347880453, 0.0257016741670668, 0.553668140666559, 0.231929493369535, 0.16080556740053, 0.328476255293936, 0.889088838826865, 0.526298201177269, 0.166863136924803, 0.707592055667192, 0.948931826511398, 0.396013859659433, 0.560890865977854, 0.138961325865239, 0.873485652264208, 0.803393356269225, 0.74370941426605, 0.045386171201244, 0.101739597739652, 0.374952184502035, 0.156885433709249, 0.67593101086095, 0.586169467540458, 0.977921136189252, 0.689145314041525, 0.840733338613063, 0.315945074660704, 0.797567360568792, 0.138057034695521, 0.360600348562002, 0.62541793868877, 0.406335793668404, 0.273074768716469, 0.405932951020077, 0.28534838068299, 0.768955264706165, 0.614511366467923, 0.180820910725743, 0.109653240302578, 0.0922589576803148, 0.315350506221876, 0.984867434250191, 0.186281649861485, 0.98500760528259, 0.687679821392521, 0.888892240822315, 0.512016676133499, 0.613913289271295, 0.78358856565319, 0.47354216943495, 0.202153110178187, 0.16619572439231, 0.529476801166311, 0.288211708888412, 0.660311228828505, 0.404482398880646, 0.512644101167098, 0.445312301395461, 0.645701993023977, 0.460568506736308, 0.668555311392993, 0.266327602788806, 0.0984652736224234, 0.739839452551678, 0.827053441200405, 0.143746328772977, 0.187439226778224, 0.735406500054523, 0.505220007151365, 0.295801528729498, 0.903635792667046, 0.756054609548301, 0.945119580021128, 0.350051714107394, 0.745320975314826, 0.131606067996472, 0.726553367683664, 0.59209707705304, 0.873026854824275, 0.561838852940127, 0.868740631267428, 0.0801749548409134, 0.103819658979774, 0.640791645273566, 0.134204810950905, 0.196401910856366, 0.316887109773234, 0.285753283882514, 0.271389214321971, 0.721793426200747, 0.981942146783695, 0.119174744002521, 0.371819758787751, 0.526617814321071, 0.705109936418012, 0.291205396875739, 0.813961637439206, 0.70934584736824, 0.327324705896899, 0.749533875146881, 0.682024102658033, 0.758025788003579, 0.589921600418165, 0.201801169896498, 0.187780044041574, 0.403755306266248, 0.144284036476165, 0.298606186173856, 0.132937862537801, 0.219614737667143, 0.894798097666353, 0.137923413654789, 0.321002187673002, 0.723915118025616, 0.565477153286338, 0.781451318180189, 0.434031368698925, 0.707067483570427, 0.038917622063309, 0.662090173456818, 0.73037575581111, 0.59144220338203, 0.924770575016737, 0.595141604542732, 0.282344592968002, 0.24166904296726, 0.280197930056602, 0.879445380298421, 0.978330681566149, 0.168315048096702, 0.812209651106969, 0.9008061690256, 0.564602505881339, 0.576646602479741, 0.773620571708307, 0.309505322482437, 0.481889286777005, 0.0243766773492098, 0.0607214251067489, 0.232588523067534, 0.468195904744789, 0.0788588826544583, 0.498393051559106, 0.940175841795281, 0.562683392083272, 0.680497571127489, 0.000745261553674936, 0.611736243590713, 0.404788655927405, 0.591872378252447, 0.117165075615048, 0.0192820732481778, 0.588208245346323, 0.667417507618666, 0.731040968792513, 0.140573708806187, 0.67832441511564, 0.69255167641677, 0.0437985460739583, 0.533890627091751, 0.379914657445624, 0.00858599343337119, 0.816253013908863, 0.295864444691688, 0.122048845980316, 0.0940997246652842, 0.650272171013057, 0.0553212456870824, 0.64778950670734, 0.952510003466159, 0.327008467633277, 0.607657962013036, 0.860665498767048, 0.0310798613354564, 0.161006239475682, 0.431680656038225, 0.339002506574616, 0.314627009676769, 0.845794194610789, 0.498776492895558, 0.213816969655454, 0.861846572021022, 0.302772739902139, 0.923288057558239, 0.579306723782793, 0.984853580128402, 0.944155388046056, 0.172454789746553, 0.945735058514401, 0.565786353312433, 0.851248633116484, 0.385770459193736, 0.25190519914031, 0.244709004415199, 0.752816929947585, 0.592083170544356, 0.598646000027657, 0.226413058349863, 0.33250065613538, 0.183948422083631, 0.19166781520471, 0.419709642184898, 0.420224759494886, 0.0866960522253066, 0.0871822524350137, 0.159378487616777, 0.971111221238971, 0.488832448376343, 0.764163428219035, 0.759997247252613, 0.975126592442393, 0.47841233597137, 0.12810812308453, 0.757524422835559, 0.908468778245151, 0.61426290939562, 0.854780389694497, 0.0726765680592507, 0.782641697209328, 0.100476315710694, 0.732006269274279, 0.42648012121208, 0.215608597965911, 0.123775037936866, 0.94612500583753, 0.0387559644877911, 0.00363942445255816, 0.904529931489378, 0.708756293170154, 0.379111863207072, 0.180028842762113, 0.819711116142571, 0.407481726491824, 0.366289080586284, 0.562746931565925, 0.633583943592384, 0.222680744482204, 0.00439866608940065, 0.179265508893877, 0.163709605811164, 0.332878219662234, 0.386155645363033, 0.0412448972929269, 0.636173987528309, 0.463790910551324, 0.898313015699387, 0.296407845336944, 0.388555900892243, 0.506861018482596, 0.393245574319735, 0.140589318005368, 0.413680299418047, 0.895250096451491, 0.0273504622746259, 0.390289203263819, 0.114639557432383, 0.125631371513009, 0.0473289131186903, 0.500197143992409, 0.335173290688545, 0.475437413435429, 0.305881184292957, 0.671624513110146, 0.78435065667145, 0.586173528572544, 0.859205979155377, 0.91187102929689, 0.171516240108758, 0.568387421779335, 0.329383664764464, 0.764269259758294, 0.162075772183016, 0.0917558800429106, 0.498564831679687, 0.671849261038005, 0.422837830148637, 0.880098348250613, 0.17486301693134, 0.722639271756634, 0.195324061904103, 0.168918929062784, 0.411378992488608, 0.25138857588172, 0.598465153947473, 0.312840997474268, 0.127453058026731, 0.483073776820675, 0.0694796864409, 0.414187713060528, 0.130995168117806, 0.855844051577151, 0.465336777968332, 0.863483853172511, 0.788173291133717, 0.986717264633626, 0.623459125403315, 0.51597937149927, 0.172988495556638, 0.66728431545198, 0.850547660375014, 0.603936887579039, 0.31910399463959, 0.45884582423605, 0.432626772904769, 0.230686279479414, 0.740641543874517, 0.983171084662899, 0.82874444569461, 0.56849360303022, 0.520462475018576, 0.331214098026976, 0.21142575214617, 0.502788838697597, 0.944401532178745, 0.530308772344142, 0.339650472858921, 0.47824402153492, 0.887685711262748, 0.659981111064553, 0.921085347188637, 0.810822007944807, 0.491516785230488, 0.0807979714591056, 0.236390282399952, 0.279511799104512, 0.619145791977644, 0.560750063043088, 0.0877040692139417, 0.694293829379603, 0.0303789267782122, 0.262588931247592, 0.378934751264751, 0.747180519392714, 0.376285756705329, 0.950430573895574, 0.000592660857364535, 0.232558073010296, 0.948537984862924, 0.975097294198349, 0.229444032534957, 0.434831802034751, 0.067268836311996, 0.147899334551767, 0.272310059284791, 0.524042938370258, 0.810155276674777, 0.976215238450095, 0.876013100612909, 0.987491950625554, 0.838664473965764, 0.290434236172587, 0.471019899705425, 0.220580665394664, 0.653011447517201, 0.302660559304059, 0.672078305622563, 0.561858527129516, 0.158962392481044, 0.394874496618286, 0.296012638369575, 0.587747772457078, 0.923173048533499, 0.397824502550066, 0.828210402280092, 0.117696603760123, 0.267516901250929, 0.636213528458029, 0.754209423670545, 0.356001936364919, 0.230784714920446, 0.681864386657253, 0.610161615302786, 0.527645993977785, 0.0164591101929545, 0.152065603761002, 0.461592234671116, 0.318451947299764, 0.404893778730184, 0.540631148032844, 0.185235136654228, 0.392055868171155, 0.0885023446753621, 0.520185227505863, 0.578731587156653, 0.524538907688111, 0.881607243558392, 0.131209205836058, 0.488581941695884, 0.154642744222656, 0.0265669869259, 0.996991156134754, 0.834439736558124, 0.520977677078918, 0.150256387889385, 0.761117015732452, 0.350164999254048, 0.428665673593059, 0.0560455336235464, 0.310032770968974, 0.0610715651419014, 0.871281517203897, 0.184450182365254, 0.736488445661962, 0.756285875337198, 0.277768790489063, 0.566505039576441, 0.896246418589726, 0.95853450614959, 0.874750504503027, 0.0168551784008741, 0.816942626144737, 0.525022267596796, 0.177697058999911, 0.602534285979345, 0.347573277540505, 0.265022631036118, 0.617054169997573, 0.732607366750017, 0.749757421668619, 0.546940963482484, 0.288561792112887, 0.360019457293674, 0.0201202321331948, 0.381626260699704, 0.442455257056281, 0.079112617764622, 0.228238570736721, 0.322412659414113, 0.673494764836505, 0.895470133516937, 0.376892961096019, 0.676557713421062, 0.233604412525892, 0.329586202977225, 0.153539561666548, 0.73335594823584, 0.644878488732502, 0.867889978224412, 0.711359289241955, 0.214133645873517, 0.209836343070492, 0.0356906414963305, 0.192679834552109, 0.979703089222312, 0.262893429026008, 0.4316752392333, 0.0646732416935265, 0.615872852969915, 0.918116956949234, 0.847007717937231, 0.971745846793056, 0.00801155855879188, 0.923083682078868, 0.178841550368816, 0.0630662215407938, 0.167766096303239, 0.978924085618928, 0.875845798756927, 0.74003918748349, 0.676606735913083, 0.872477106750011, 0.561448477208614, 0.710860214428976, 0.775917960330844, 0.895800209138542, 0.717388630611822, 0.106744910124689, 0.417166819097474, 0.720758825307712, 0.0927380549255759, 0.546242141863331, 0.770855001639575, 0.0217703629750758, 0.890302506973967, 0.154255953617394, 0.399082737974823, 0.625154075212777, 0.866359584499151, 0.207459912402555, 0.0528666030149907, 0.714785255724564, 0.479803605936468, 0.249940327368677, 0.964672264177352, 0.799881111364812, 0.479371314169839, 0.0443333419971168, 0.799163210205734, 0.719540245598182, 0.707329176366329, 0.543197964085266, 0.0865204487927258, 0.245725208893418, 0.289947208715603, 0.369423666736111, 0.653847910230979, 0.217003552941605, 0.47728944895789, 0.581432088045403, 0.856342976912856, 0.141315585235134, 0.781053911428899, 0.782447389792651, 0.697796673048288, 0.975292314775288, 0.373759947018698, 0.437863871688023, 0.207444449188188, 0.932642517844215, 0.216318636666983, 0.422740825684741, 0.00624043238349259, 0.959941376000643, 0.69157274486497, 0.18601042823866, 0.961711262352765, 0.687899705488235, 0.363955528708175, 0.189117638627067, 0.26907165418379, 0.892366836313158, 0.643706730799749, 0.785383344860747, 0.947996832896024, 0.0398137483280152, 0.537790692411363, 0.111443746136501, 0.723794379737228, 0.985559830442071, 0.84933021850884, 0.0687404698692262, 0.841466228244826, 0.200990235898644, 0.385261445539072, 0.977512510726228, 0.0422723647207022, 0.102149070240557, 0.874700621934608, 0.1049673196394, 0.356974135152996, 0.441480020992458, 0.976180779747665, 0.156607088632882, 0.323326140409335, 0.422505279071629, 0.664734658552334, 0.571023928234354, 0.411075433716178, 0.302728943526745, 0.489540531532839, 0.157278725178912, 0.991375125944614, 0.92671130457893, 0.812214175006375, 0.322537840111181, 0.168678259011358, 0.554761177161708, 0.655319227837026, 0.208400727016851, 0.074362589744851, 0.125103488797322, 0.984174585901201, 0.591437845258042, 0.00507185724563897, 0.0965164804365486, 0.019614000339061, 0.0731614348478615, 0.37506387126632, 0.104424218181521, 0.324166625505313, 0.186517084715888, 0.335062138037756, 0.181418574647978, 0.509644304635003, 0.706710108090192, 0.060058414703235, 0.459590561687946, 0.838081233901903, 0.865278931567445, 0.315370313357562, 0.0575890967156738, 0.904992730356753, 0.879326300229877, 0.564955997513607, 0.148839666042477, 0.929143688175827, 0.429725867463276, 0.426099132047966, 0.00728612300008535, 0.460097413277254, 0.702191048068926, 0.0578466278966516, 0.795599594479427, 0.908823024481535, 0.647009758045897, 0.866070275194943, 0.295793359400705, 0.615081968950108, 0.277325787348673, 0.741083175642416, 0.803613391937688, 0.498588484944776, 0.372300931252539, 0.149931589374319, 0.344307675026357, 0.260766243794933, 0.513775942381471, 0.639587186975405, 0.930070370202884, 0.991040745517239, 0.693391385488212, 0.76254046917893, 0.293650425970554, 0.033340489724651, 0.633523118915036, 0.11063348967582, 0.622520602773875, 0.382067410042509, 0.364967240486294, 0.592676289612427, 0.645128910895437, 0.932243133196607, 0.682968755019829, 0.814869384979829, 0.538172542816028, 0.892137321410701, 0.944793654838577, 0.192694402066991, 0.22827699012123, 0.882343719247729, 0.538190169492736, 0.626705530798063, 0.860741224372759, 0.394213985418901, 0.0402473998256028, 0.455033849226311, 0.675020781112835, 0.945210566744208, 0.794961860170588, 0.745565022109076, 0.0730010587722063, 0.854125570738688, 0.830417464720085, 0.392569426679984, 0.539286464685574, 0.973610946442932, 0.238613706314936, 0.597584093920887, 0.202141467714682, 0.549157851142809, 0.484516033437103, 0.204168861964718, 0.351019570138305, 0.718884598929435, 0.787767331115901, 0.0483347875997424, 0.156574832974002, 0.893831558758393, 0.945169688668102, 0.522170349024236, 0.380111966747791, 0.257205907022581, 0.66545895440504, 0.829317881725729, 0.592208539834246, 0.494943951023743, 0.335144028067589, 0.676316779106855, 0.908349787816405, 0.668312439695001, 0.942235197173432, 0.164983129128814, 0.535427201539278, 0.647654585773125, 0.752435371046886, 0.852146129589528, 0.158131476026028, 0.355523865437135, 0.831168810836971, 0.178531744284555), actual = c(0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.987465495476499, 0.869082706514746, 0.932040317216888, 0.845160360448062, 0.71265245298855, 0.932613522512838, 0.331665740581229, 0.240069342544302, 0.276143228868023, 0.310506004374474, 0.85169758182019, 0.465620917733759, 0.639308216283098, 0.126344448653981, 0.094157921615988, 0.580156250391155, 0.790483116405085, 0.093159053940326, 0.575649632373825, 0.200663009658456, 0.374691642588004, 0.420440977904946, 0.97352374298498, 0.373564104782417, 0.475291253998876, 0.407490174286067, 0.0468157436698675, 0.483216615160927, 0.0188193945214152, 0.238004809478298, 0.694137287326157, 0.307686327490956, 0.5858711944893, 0.813257604138926, 0.843576977960765, 0.985131677705795, 0.0713271608110517, 0.683301741024479, 0.465929335914552, 0.392818154534325, 0.259295320836827, 0.843209486454725, 0.933512115618214, 0.590763257816434, 0.184166153892875, 0.924425572156906, 0.507302536861971, 0.824926323257387, 0.260026248637587, 0.689540814608335, 0.184835343621671, 0.973116444889456, 0.28147906367667, 0.0393705377355218, 0.906643446534872, 0.329619252821431, 0.0829287422820926, 0.391820452408865, 0.431244901847094, 0.593193437904119, 0.897079500369728, 0.387192913098261, 0.475297632627189, 0.814030883368105, 0.648924659471959, 0.146086602704599, 0.983931219903752, 0.478893944527954, 0.379284022375941, 0.304885892895982, 0.533306668046862, 0.199316843412817, 0.628257194068283, 0.416492170887068, 0.577436323044822, 0.394435623195022, 0.842447337228805, 0.377742970595136, 0.366262624971569, 0.390811317134649, 0.0771960238926113, 0.359044829616323, 0.417241687886417, 0.736265804618597, 0.76419591368176, 0.629031223943457, 0.947051457827911, 0.428875378565863, 0.394586261129007, 0.339718399103731, 0.174877021461725, 0.0437046671286225, 0.395849468419328, 0.282478442415595, 0.0796212854329497, 0.115072175161913, 0.624994293320924, 0.871335068484768, 0.0390487180557102, 0.814349903026596, 0.436411695089191, 0.0909513782244176, 0.482608532300219, 0.516312994994223, 0.445487776538357, 0.352507145376876, 0.760184616548941, 0.475239026593044, 0.621521047549322, 0.996184831485152, 0.678164733340964, 0.171743258135393, 0.284876419464126, 0.0583440754562616, 0.0947308368049562, 0.0723494794219732, 0.961345048621297, 0.589776404667646, 0.307602266548201, 0.238569046836346, 0.196730875177309, 0.330153145827353, 0.0459982855245471, 0.509137634886429, 0.346019740914926, 0.411105158738792, 0.260328492848203, 0.0194898070767522, 0.0530242533423007, 0.739037629682571, 0.633929467760026, 0.347178646828979, 0.893051369348541, 0.412810190580785, 0.0369224594905972, 0.761230314150453, 0.255939054768533, 0.610770509578288, 0.977534647099674, 0.979641588171944, 0.313109531067312, 0.15536181954667, 0.073183138621971, 0.07572455634363, 0.192434794735163, 0.30731773446314, 0.451876300154254, 0.604198052780703, 0.00259090284816921, 0.85069015645422, 0.210357223637402, 0.946733386255801, 0.704857836710289, 0.801716211019084, 0.33902655239217, 0.627822375390679, 0.15568850026466, 0.0295827228110284, 0.245844915509224, 0.569102995796129, 0.767114343354478, 0.644230683799833, 0.180243021808565, 0.797458542510867, 0.348968075588346, 0.23616861552, 0.652188249863684, 0.597856737906113, 0.000586424954235554, 0.824791789287701, 0.827349125873297, 0.908357321983203, 0.417781615629792, 0.844240715261549, 0.413424735423177, 0.336003873497248, 0.428009348222986, 0.759384911274537, 0.4382642544806, 0.368830217747018, 0.23341527627781, 0.875795257044956, 0.00044035934843123, 0.859296499751508, 0.659792319638655, 0.307941843289882, 0.126508179120719, 0.261683857999742, 0.872206127271056, 0.192697110585868, 0.264960241271183, 0.972546242177486, 0.0335854387376457, 0.814476030413061, 0.852353771682829, 0.82682118890807, 0.957660445710644, 0.840256746159866, 0.355795679148287, 0.189507109113038, 0.573321983451024, 0.0404413128271699, 0.301373338559642, 0.809738088632002, 0.503114535938948, 0.192738963291049, 0.874975014710799, 0.176143399439752, 0.22022628551349, 0.0188939284998924, 0.528014652198181, 0.921293846564367, 0.355606106575578, 0.0771478698588908, 0.846779506187886, 0.282040243968368, 0.348013538867235, 0.120685520814732, 0.525566417723894, 0.167898697312921, 0.182183586992323, 0.984085082542151, 0.460761385038495, 0.0967759746126831, 0.538412474328652, 0.970859876368195, 0.0658998836297542, 0.628666270989925, 0.579228611895815, 0.887258485658094, 0.80083370138891, 0.0120685733854771, 0.918068280210719, 0.609372907318175, 0.602660091593862, 0.994752456434071, 0.573385550407693, 0.829428710974753, 0.62869614828378, 0.722668728791177, 0.553592358250171, 0.453467167448252, 0.0766509326640517, 0.392170348670334, 0.452094666194171, 0.0966799892485142, 0.68007015879266, 0.342340956442058, 0.0587837605271488, 0.475263032130897, 0.690058881649747, 0.986854551592842, 0.335438538342714, 0.796322465175763, 0.591713289963081, 0.375980156240985, 0.577915346715599, 0.815166373504326, 0.906825445126742, 0.921635534614325, 0.328718113247305, 0.143776735756546, 0.0239904031623155, 0.0585775165818632, 0.989989358233288, 0.723112700274214, 0.488896063296124, 0.733111509354785, 0.233634324045852, 0.169795929687098, 0.465334338136017, 0.895970877725631, 0.775493578054011, 0.973187691997737, 0.862685927422717, 0.210380653617904, 0.351019817171618, 0.862351757241413, 0.997616265900433, 0.363905722042546, 0.243671489180997, 0.461178167257458, 0.299948294879869, 0.320533286780119, 0.602506297640502, 0.244714644970372, 0.75884315255098, 0.0416624636854976, 0.39232699945569, 0.357446643756703, 0.0795784702058882, 0.205901093315333, 0.770897717913613, 0.268178412923589, 0.593294192105532, 0.999467013636604, 0.40153193497099, 0.791328636696562, 0.52508883154951, 0.41683000465855, 0.820416906615719, 0.860786315286532, 0.288724497193471, 0.0810809186659753, 0.811325424583629, 0.69604986323975, 0.744830637006089, 0.61618689680472, 0.948849348118529, 0.657062451122329, 0.989127125358209, 0.0506398943252861, 0.314069087151438, 0.788734546862543, 0.589068731525913, 0.442379041342065, 0.87037993548438, 0.348201089771464, 0.613189413677901, 0.628027595113963, 0.480027477722615, 0.257992068072781, 0.855080239474773, 0.136832752497867, 0.645708682946861, 0.797733307350427, 0.597286659525707, 0.695935048628598, 0.667182968463749, 0.384346510283649, 0.71032521687448, 0.320400838041678, 0.150760079035535, 0.498444145545363, 0.76005207747221, 0.239780227420852, 0.59013365348801, 0.856970452005044, 0.711404267232865, 0.965763933723792, 0.998249166412279, 0.125567147508264, 0.535178613616154, 0.139978002756834, 0.531524279387668, 0.168608350446448, 0.267564342124388, 0.847413421841338, 0.519389037974179, 0.208305661799386, 0.833036828320473, 0.769844490801916, 0.793815214186907, 0.219923595897853, 0.392812328413129, 0.981744457269087, 0.510906299343333, 0.188867998542264, 0.225914619863033, 0.0974287742283195, 0.977507416857406, 0.346538490848616, 0.317398415412754, 0.897706864867359, 0.227148357545957, 0.948101110756397, 0.140932220965624, 0.146638586185873, 0.191346798092127, 0.902503099991009, 0.537898792885244, 0.547244013752788, 0.202184143476188, 0.930705455131829, 0.997031525243074, 0.673524042824283, 0.257630655774847, 0.892267632763833, 0.380140777444467, 0.870160866994411, 0.381591843441129, 0.180419939570129, 0.258763251593336, 0.254692712798715, 0.539130014600232, 0.201607055030763, 0.404982677660882, 0.643477459670976, 0.389331557555124, 0.391148639144376, 0.317644411930814, 0.236624739132822, 0.404721214203164, 0.791742025874555, 0.00518240942619741, 0.409710677806288, 0.0699021264445037, 0.0633501624688506, 0.826137477299199, 0.81922788056545, 0.790467098820955, 0.700706772273406, 0.56201135693118, 0.223297434858978, 0.754058932419866, 0.98518061847426, 0.473477358696982, 0.784382649930194, 0.801675355993211, 0.797339589335024, 0.459199842996895, 0.335826022783294, 0.0472537574823946, 0.155283414758742, 0.496504644630477, 0.81036616046913, 0.819606236414984, 0.385493346257135, 0.919138334924355, 0.87927311193198, 0.813397443387657, 0.0743251016829163, 0.00237232353538275, 0.601775022456422, 0.649301081662998, 0.876992091536522, 0.629309015348554, 0.532468796474859, 0.892235762672499, 0.417645909357816, 0.827656391309574, 0.609603715827689, 0.930838252650574, 0.063318622065708, 0.815383509267122, 0.5813568320591, 0.759720085188746, 0.738799675367773, 0.313271159306169, 0.609720684122294, 0.521156832342967, 0.526199354790151, 0.633071853546426, 0.856423564488068, 0.813453862909228, 0.0529748187400401, 0.215240070829168, 0.359200758393854, 0.987782961223274, 0.949023905675858, 0.168928673025221, 0.412106862524524, 0.361301906406879, 0.906225743470713, 0.966311274562031, 0.512813809793442, 0.457580233924091, 0.309328796342015, 0.342805510619655, 0.76818799530156, 0.370723149506375, 0.940222035394982, 0.228485420579091, 0.519942530663684, 0.156035184627399, 0.203132370486856, 0.999848301289603, 0.968080979771912, 0.725547352340072, 0.753765563713387, 0.434842768125236, 0.749425149988383, 0.701948308385909, 0.0654223009478301, 0.0848557553254068, 0.903183133574203, 0.732719885883853, 0.817765790037811, 0.414458766812459, 0.264793621143326, 0.861509087728336, 0.180063413223252, 0.593545068753883, 0.0606375157367438, 0.920239473925903, 0.787408847361803, 0.912427448667586, 0.430917552905157, 0.399009634274989, 0.672643065219745, 0.611972565297037, 0.576926567126065, 0.191229055868462, 0.12046475126408, 0.500118974829093, 0.560098400572315, 0.158553897403181, 0.131492435233667, 0.652630011551082, 0.958281560568139, 0.682835318846628, 0.442332198377699, 0.26572457770817, 0.332861980656162, 0.272012251196429, 0.244514248101041, 0.596581688383594, 0.701699278084561, 0.74572089780122, 0.693480850197375, 0.645124037982896, 0.30549998418428, 0.561379763530567, 0.31171609624289, 0.153918475843966, 0.64862407441251, 0.717248307308182, 0.126081517897546, 0.066427601268515, 0.746973665431142, 0.503518328536302, 0.99126001726836, 0.873935153242201, 0.893360634800047, 0.648203402524814, 0.151704907417297, 0.433202091604471, 0.139316213782877, 0.53008532547392, 0.0791535482276231, 0.504095549229532, 0.991502135060728, 0.231780646368861, 0.170910652726889, 0.475919116521254, 0.471484306268394, 0.0587207018397748, 0.440301653929055, 0.538460691226646, 0.344865432241932, 0.0834099452476948, 0.75830847513862, 0.154259445378557, 0.857562416931614, 0.735447933664545, 0.111992012476549, 0.493449714267626, 0.643133053788915, 0.0347801009193063, 0.269331355812028, 0.251248000655323, 0.170252400683239, 0.331804502988234, 0.258762452751398, 0.095087479101494, 0.698359298054129, 0.562377512920648, 0.296853594249114, 0.99380100867711, 0.60017270850949, 0.978168642148376, 0.207336693769321, 0.572134715039283, 0.648864955408499, 0.0328151895664632, 0.0646670591086149, 0.109830925241113, 0.201123085338622, 0.551652864553034, 0.28802526416257, 0.926304179709405, 0.0539754149504006, 0.977377896429971, 0.338131151394919, 0.594161886489019, 0.396914296085015, 0.286183075513691, 0.563302700407803, 0.176870639435947, 0.840410133358091, 0.94739226391539, 0.748721778392792, 0.393772668438032, 0.0126319820992649, 0.289751100819558, 0.83906252332963, 0.0653540331404656, 0.197569498093799, 0.996236223960295, 0.341497301356867, 0.0602973096538335, 0.920828240457922, 0.065213605761528, 0.154777283547446, 0.660603718133643, 0.60465403040871, 0.251901966053993, 0.83688419079408, 0.691878472454846, 0.764465728309005, 0.256616664351895, 0.451563101727515, 0.264791894936934, 0.383855685358867, 0.383869261015207, 0.535481558647007, 0.535007770638913, 0.536514373961836, 0.600772280013189, 0.9054543687962, 0.344177070772275, 0.757877563359216, 0.970049845520407, 0.638550478033721, 0.223775721620768, 0.750712880631909, 0.290057536913082, 0.260555147659034, 0.259921498596668, 0.0993728491012007, 0.152333854930475, 0.321194605901837, 0.986751165241003, 0.573663784191012, 0.659104860387743, 0.128656554035842, 0.753427412360907, 0.894827816169709, 0.519819832406938, 0.663282563211396, 0.559806177625433, 0.424929824890569, 0.563017296837643, 0.837808839743957, 0.752447725040838, 0.239191495813429, 0.551319691585377, 0.551026297034696, 0.86052892007865, 0.611270625842735, 0.869796113343909, 0.386675008805469, 0.983331238850951, 0.433850993867964, 0.923841584706679, 0.0152126781176776, 0.384616321884096, 0.408547475934029, 0.0452829282730818, 0.0532498750835657, 0.695471760816872, 0.899630099069327, 0.322481114882976, 0.767566903727129, 0.708459236193448, 0.153508730465546, 0.696560009615496, 0.703446021303535, 0.321310511557385, 0.0999321651179343, 0.109029718441889, 0.424749099649489, 0.918900179676712, 0.384887743042782, 0.256815223023295, 0.803859008941799, 0.821627796627581, 0.743798449169844, 0.124442412052304, 0.0488326221238822, 0.924822553060949, 0.44627572526224, 0.556253046030179, 0.0687438466120511, 0.952433867612854, 0.969140063039958, 0.175065369810909, 0.382338274968788, 0.48327290196903, 0.250292638782412, 0.00634699990041554, 0.508294138824567, 0.150319874752313, 0.788158755283803, 0.758010131539777, 0.721955293556675, 0.392507591284811, 0.786994555965066, 0.217380177928135, 0.560172734316438, 0.135879656299949, 0.700116131920367, 0.0689288647845387, 0.274768219795078, 0.891811126144603, 0.225127408280969, 0.194745239336044, 0.190429742913693, 0.76847410434857, 0.249203679617494, 0.12227028911002, 0.0168822582345456, 0.748242367524654, 0.547347766812891, 0.296844200463966, 0.292759932111949, 0.0414276078809053, 0.731525627663359, 0.348212926415727, 0.848625348880887, 0.811338881263509, 0.932927677873522, 0.701506971614435, 0.532734289765358, 0.31770720356144, 0.707292657578364, 0.402395238401368, 0.0790572620462626, 0.123998111113906, 0.652726948494092, 0.269172376487404, 0.749415871454403, 0.533491298323497, 0.729698089649901, 0.621970565291122, 0.241602637805045, 0.535861831856892, 0.694778471253812, 0.749141194857657, 0.935023482423276, 0.953345018671826, 0.134290141519159, 0.183697674423456, 0.870819847332314, 0.966166469966993, 0.886644039768726, 0.00652860454283655, 0.107173894532025, 0.829858352430165, 0.79991278774105, 0.934268277371302, 0.0222426175605506, 0.0370665178634226, 0.0162838159594685, 0.603517869021744, 0.732137501705438, 0.602755270665511, 0.851541631156579, 0.31076420051977, 0.903135822154582, 0.834955219179392, 0.512955129379407, 0.148709352593869, 0.295067804865539, 0.805309478193521, 0.679735102225095, 0.0302850543521345, 0.697545014321804, 0.646174459950998, 0.243708635447547, 0.605487684020773, 0.157111934153363, 0.55448675295338, 0.2506299756933, 0.0280767851509154, 0.797426498262212, 0.697192186256871, 0.198076590429991, 0.694352712482214, 0.248129616957158, 0.142702127806842, 0.180430534761399, 0.882033351343125, 0.574066231260076, 0.253400027286261, 0.341768813086674, 0.0760054639540613, 0.897292301757261, 0.370341703528538, 0.18481901451014, 0.519532721489668, 0.77695198263973, 0.198293180437759, 0.0484465302433819, 0.0997547628358006, 0.758892397861928, 0.605979998130351, 0.554148523835465, 0.600527115631849, 0.131079020211473, 0.912067376775667, 0.239707562839612, 0.679913043044508, 0.382359866285697, 0.338208266999573, 0.74545227480121, 0.475878769066185, 0.443429521983489, 0.440876497188583, 0.282951046945527, 0.161915732780471, 0.344163598492742, 0.895074568223208, 0.689334331080317, 0.657891011796892, 0.514373159967363, 0.191929896827787, 0.775355260353535, 0.193600415019318, 0.81738751847297, 0.231280128471553, 0.197003584820777, 0.446698842570186, 0.240930248983204, 0.715048887766898, 0.723005298757926, 0.727609873283654, 0.768311762949452, 0.849040156463161, 0.225157751468942, 0.12251822813414, 0.370391811942682, 0.806793024064973, 0.340213020099327, 0.0734420090448111, 0.605524270562455, 0.974195188377053, 0.759081557858735, 0.985084645450115, 0.507525104796514, 0.922907006461173, 0.606145780300722, 0.747554318280891, 0.153586626984179, 0.358729629777372, 0.620762144448236, 0.130875952076167, 0.741083140252158, 0.0360916312783957, 0.282617122633383, 0.111850777873769, 0.0871494968887419, 0.126417425926775, 0.263851755997166, 0.345409068046138, 0.308120260247961, 0.751034597633407, 0.623398178257048, 0.808024021796882, 0.637526290956885, 0.056016692193225, 0.779127697227523, 0.986547876382247, 0.618194553768262, 0.470958593534306, 0.15199363976717, 0.31754782423377, 0.497518310323358, 0.70631857495755, 0.816082191187888, 0.968560243956745, 0.624128879746422, 0.637137363431975, 0.20214985567145, 0.374622793169692, 0.511513003380969, 0.864040918415412, 0.402089164825156, 0.898599609965459, 0.912723386194557, 0.450476833619177, 0.458297440782189, 0.449016322614625, 0.39355145720765, 0.861137885833159, 0.341970170848072, 0.975022561149672, 0.934739372925833, 0.00479455618187785, 0.653104649623856, 0.863674734020606, 0.543829549103975, 0.787559141172096, 0.539294228423387, 0.745276517001912, 0.67737101088278, 0.163386140251532, 0.631371690426022, 0.412612956482917, 0.451496269088238, 0.883386552566662, 0.736549082910642, 0.731654093135148, 0.442849051440135, 0.0463878230657429, 0.861523543484509, 0.849392648087814, 0.348950555780903, 0.626308890292421, 0.0609097711276263, 0.0372228962369263, 0.0422425193246454, 0.77346029621549, 0.64774476364255, 0.294084617169574, 0.375115546165034, 0.3495752888266, 0.961989574832842, 0.108640572987497, 0.388928014086559, 0.762221638346091, 0.650095388526097, 0.730982954846695, 0.571764377877116, 0.440816838294268, 0.820507612312213, 0.736890138592571, 0.125726534985006, 0.092276587150991, 0.158259357558563, 0.425152198877186, 0.583753718063235, 0.431687173666432, 0.88734235544689, 0.469037733972073, 0.221677489345893, 0.0499317732173949, 0.0374060103204101, 0.300910050049424, 0.1306421039626, 0.0894392065238208, 0.376086908392608, 0.707246421603486, 0.647849361412227, 0.954508520197123, 0.321606656303629, 0.574518207460642, 0.307334369746968, 0.104686220642179, 0.0159443675074726, 0.732323799282312, 0.851860921829939, 0.30109992204234, 0.537931737257168, 0.597750519867986, 0.241769860265777, 0.152520907809958, 0.664363193092868, 0.37837175745517, 0.358454620465636, 0.230120087508112, 0.136595056625083, 0.740128278499469, 0.49125993042253, 0.847073641838506, 0.884631851222366, 0.202917483402416, 0.890476528787985, 0.143223253078759, 0.554235451156273, 0.382339563686401, 0.294006870593876, 0.769020779523998, 0.862923075910658, 0.605265820864588, 0.828775164205581, 0.034109617350623, 0.189963551936671, 0.975091341650113, 0.985872811172158, 0.458237475715578, 0.0698088360950351, 0.776876645162702, 0.0065232461784035, 0.191844961605966, 0.200150785269216, 0.233874598750845, 0.772215465083718, 0.883940459461883, 0.177645612275228, 0.579113939311355, 0.523425500607118, 0.474796182708815, 0.883859753608704, 0.785043488023803, 0.178687990410253, 0.273233251646161, 0.433491979725659, 0.517329991096631, 0.476311082718894, 0.348286856431514, 0.905165059026331, 0.00990166538394988, 0.256151049397886, 0.0358424226287752, 0.672250490868464, 0.865856058429927, 0.90115938638337, 0.282423477619886, 0.252629918279126), actual = c(0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"