Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.2880798        0.2880798 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5390885        0.5390885        0.5393150 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.498086019884795, 0.130652618827298, 0.772992964135483, 0.8877342347987, 0.469386268872768, 0.863599934848025, 0.854750119848177, 0.0460708332248032, 0.197716496884823, 0.625174106098711, 0.925837553106248, 0.756443127291277, 0.619488721946254, 0.654338343068957, 0.316892498871312, 0.0326830022968352, 0.985744462581351, 0.162387888180092, 0.683840416837484, 0.435383186908439, 0.981846852693707, 0.380300427321345, 0.733481465373188, 0.224523628829047, 0.711421004962176, 0.314431596780196, 0.977315093856305, 0.415084619307891, 0.484608904458582, 0.514734360855073, 0.612819891655818, 0.611540474696085, 0.824569167103618, 0.249213212868199, 0.759572825860232, 0.271031518001109, 0.280566910281777, 0.645885101985186, 0.498008598573506, 0.639472719747573, 0.759019579039887, 0.224214439280331, 0.598661831347272, 0.412980691995472, 0.645222190301865, 0.353987063746899, 0.402372244047001, 0.371591631090268, 0.308438840322196, 0.186473980778828, 0.503264374798164, 0.754312153207138, 0.699096489930525, 0.674039661884308, 0.112581304972991, 0.325816204538569, 0.868215198162943, 0.820027291076258, 0.346607758896425, 0.57010551309213, 0.373503796290606, 0.0881268926896155, 0.896861877758056, 0.200477508362383, 0.409800628200173, 0.606132147135213, 0.591164969373494, 0.0261673836503178, 0.954851179616526, 0.755357041256502, 0.274602500256151, 0.111003260593861, 0.316719641676173, 0.452777507947758, 0.970490839565173, 0.062502681510523, 0.988167101982981, 0.903906747698784, 0.213945827214047, 0.492839875398204, 0.501753846649081, 0.871819205582142, 0.692544065415859, 0.132486185990274, 0.159640456084162, 0.715194564778358, 0.259134945925325, 0.864655876532197, 0.601570978295058, 0.891813002992421, 0.0774811957962811, 0.770718916784972, 0.920624318998307, 0.722621292807162, 0.66899247164838, 0.775702005019411, 0.420771667500958, 0.400153995258734, 0.734339796472341, 0.706972183892503, 0.869992529507726, 0.274457802297547, 0.925132238538936, 0.835231536766514, 0.37060918752104, 0.190778290852904, 0.288051138399169, 0.50563762919046, 0.295068077510223, 0.226333251688629, 0.654844498028979, 0.703082600142807, 0.970991365378723, 0.794734540628269, 0.147952512837946, 0.904013059800491, 0.0486768728587776, 0.762098304461688, 0.595833278959617, 0.616913715610281, 0.687579946359619, 0.91712433937937, 0.972723553422838, 0.727373634232208, 0.84545971499756, 0.722756715025753, 0.44277670327574, 0.368638592772186, 0.68551841750741, 0.897960321279243, 0.637026581680402, 0.210659720236436, 0.321023412747309, 0.644213155843318, 0.857610783539712, 0.835582070285454, 0.334361989982426, 0.893854742869735, 0.0202918299473822, 0.395427286624908, 0.74864030117169, 0.365099249174818, 0.125645999563858, 0.0688533536158502, 0.844753880752251, 0.0974902492016554, 0.973958138842136, 0.047017972683534, 0.792236427310854, 0.869009869871661, 0.818871641997248, 0.521414010087028, 0.750180823495612, 0.544999259291217, 0.637443381128833, 0.573217462049797, 0.318026863504201, 0.852307445369661, 0.581609559012577, 0.760660720523447, 0.875340400030836, 0.0835060940589756, 0.263298879610375, 0.891597929410636, 0.248893279815093, 0.770683387061581, 0.976013279985636, 0.297235099365935, 0.0312107903882861, 0.525730324909091, 0.831366646802053, 0.339020420098677, 0.705891155870631, 0.423830499639735, 0.618959543062374, 0.00797156407497823, 0.586430728901178, 0.278348945546895, 0.460624057566747, 0.894924465566874, 0.589490090264007, 0.916186554823071, 0.696148914983496, 0.97602008539252, 0.194277659757063, 0.505566690349951, 0.00683894869871438, 0.2970693805255, 0.253269292181358, 0.624041758710518, 0.690380739048123, 0.170314932242036, 0.438700237078592, 0.210899374913424, 0.06540552014485, 0.311140873003751, 0.847193142632023, 0.2608403977938, 0.281204239698127, 0.375530230812728, 0.305539692752063, 0.624048817204311, 0.328800990246236, 0.597589979181066, 0.289642550516874, 0.955514795845374, 0.284627255285159, 0.903561087325215, 0.617151984712109, 0.316043245838955, 0.772608713479713, 0.404375371523201, 0.940555886132643, 0.555067505221814, 0.189121787203476, 0.966800678055733, 0.496802007313818, 0.92565445555374, 0.0881075283978134, 0.898390553193167, 0.235270251752809, 0.950874748872593, 0.513741568429396, 0.637787038227543, 0.196890135295689, 0.0483954227529466, 0.445203971117735, 0.0500171072781086, 0.551145856035873, 0.254909890005365, 0.354154472937807, 0.848853748291731, 0.915054226526991, 0.647152453195304, 0.355205465573817, 0.216822761343792, 0.944784947438166, 0.642332730349153, 0.594177831895649, 0.472702563041821, 0.827131830854341, 0.517470583319664, 0.793910586740822, 0.078026337781921, 0.582383055239916, 0.223903492093086, 0.39301551412791, 0.0607553627341986, 0.898329827701673, 0.0607278433162719, 0.40748928184621, 0.183698358247057, 0.363237380050123, 0.668446211609989, 0.47574620670639, 0.707304230658337, 0.247018488124013, 0.817947746487334, 0.440176323056221, 0.0625419339630753, 0.351579167414457, 0.495092449244112, 0.0618447698652744, 0.830210478277877, 0.516211911803111, 0.0366409793496132, 0.864545718301088, 0.208968237740919, 0.484782470390201, 0.77661959361285, 0.27632554830052, 0.735040363389999, 0.589705578051507, 0.849161823512986, 0.968372497474775, 0.324449157807976, 0.817879193695262, 0.643218586686999, 0.802227230975404, 0.899824614869431, 0.312872343463823, 0.616024714196101, 0.505907042417675, 0.474042990244925, 0.145885148784146, 0.105803887825459, 0.812934915302321, 0.487668211571872, 0.361527486005798, 0.0144863855093718, 0.63171745929867, 0.856701303273439, 0.87990963133052, 0.112734805326909, 0.582373837241903, 0.728917501633987, 0.0440510800108314, 0.172687663231045, 0.527340571628883, 0.779829900711775, 0.412135153543204, 0.579091730061918, 0.309791712090373, 0.0534000154584646, 0.666821589227766, 0.409493983024731, 0.540573493344709, 0.737141574034467, 0.932338570710272, 0.480901370523497, 0.545339476084337, 0.0997935989871621, 0.825319024734199, 0.778412627987564, 0.955326602095738, 0.501540453406051, 0.538376634707674, 0.933297892799601, 0.335052294656634, 0.12930977717042, 0.900138138327748, 0.927437183680013, 0.758168754400685, 0.971405776916072, 0.731609961250797, 0.648877499159425, 0.960968032013625, 0.954879274358973, 0.542624974390492, 0.0463208339642733, 0.612744045909494, 0.855174958938733, 0.00574610661715269, 0.117432512110099, 0.422147226054221, 0.814075227594003, 0.366554870968685, 0.541307487292215, 0.883070773212239, 0.0267660394310951, 0.905074249720201, 0.031406796304509, 0.00876997411251068, 0.958692545536906, 0.48906885785982, 0.0930090777110308, 0.93602950219065, 0.262813628418371, 0.80124117876403, 0.727168454555795, 0.391597371315584, 0.511284483829513, 0.741384769091383, 0.259849064750597, 0.196155926212668, 0.978401488158852, 0.567920774454251, 0.170945991994813, 0.138626436935738, 0.700782565400004, 0.561264629242942, 0.533488776301965, 0.807452168082818, 0.445441059302539, 0.18426197161898, 0.401255997829139, 0.3813636158593, 0.0343084558844566, 0.372164579574019, 0.498951745918021, 0.476833794731647, 0.517413980327547, 0.655467591714114, 0.615585979772732, 0.16671349434182, 0.936307356227189, 0.573219627374783, 0.513846016954631, 0.630880181444809, 0.295242889784276, 0.901420070091262, 0.0764207025058568, 0.783349602716044, 0.624713765457273, 0.452496705343947, 0.539133397163823, 0.651453870814294, 0.686321303248405, 0.0475609644781798, 0.437811629613861, 0.728112485958263, 0.857990461867303, 0.0219193669036031, 0.380216450663283, 0.696478451834992, 0.774575718212873, 0.180722866440192, 0.0273547780234367, 0.263870497001335, 0.376405053306371, 0.62084365892224, 0.779088425450027, 0.0592748520430177, 0.902886918047443, 0.954648672603071, 0.91621925518848, 0.0260298470966518, 0.491148357046768, 0.0988454048056155, 0.708619024371728, 0.347770652966574, 0.646267196862027, 0.837919271551073, 0.125850912882015, 0.544740717625245, 0.739647508133203, 0.115549236536026, 0.539784642634913, 0.801557207480073, 0.234935808461159, 0.483040905557573, 0.318082225508988, 0.409406767925248, 0.568479944951832, 0.435730383498594, 0.818317183991894, 0.813962686574087, 0.045771794160828, 0.477022824343294, 0.959080172004178, 0.537641848204657, 0.46039478899911, 0.709625359158963, 0.654761352343485, 0.0903292344883084, 0.196508475346491, 0.355504977982491, 0.980635756626725, 0.0394512240309268, 0.490028319880366, 0.212563580367714, 0.00294297165237367, 0.359206815715879, 0.514226655475795, 0.577654446242377, 0.277830772101879, 0.704245077911764, 0.15637658466585, 0.693642801372334, 0.154073190642521, 0.320054971845821, 0.358202878385782, 0.182561551919207, 0.546141884289682, 0.1083849738352, 0.253030264051631, 0.531317673390731, 0.900714355520904, 0.427535957191139, 0.821603571297601, 0.486839316552505, 0.255291629116982, 0.708602675702423, 0.619504197733477, 0.690710806054994, 0.334745757747442, 0.491593706887215, 0.184101898223162, 0.276658870279789, 0.826004792936146, 0.365151618607342, 0.672392188804224, 0.8888312545605, 0.650267209857702, 0.938512937398627, 0.131945971166715, 0.509876568801701, 0.677204547217116, 0.451407358516008, 0.403227230068296, 0.637381576001644, 0.144411831861362, 0.0284241293556988, 0.0275712008588016, 0.648687834385782, 0.36972021497786, 0.447116451105103, 0.192502469057217, 0.0895049285609275, 0.73343469761312, 0.0552855022251606, 0.491299316287041, 0.518241673242301, 0.368138215970248, 0.277839644579217, 0.490445513045415, 0.0857688870746642, 0.449024008819833, 0.414142307825387, 0.307039825245738, 0.557512807659805, 0.153525918722153, 0.970368671230972, 0.367762954439968, 0.492206252645701, 0.669917047489434, 0.653102758107707, 0.955136702861637, 0.321640267269686, 0.077813538024202, 0.389961696462706, 0.172333934344351, 0.447730927262455, 0.392831244971603, 0.679046698380262, 0.338062384398654, 0.106812514830381, 0.753833349794149, 0.941676063928753, 0.0715083512477577, 0.940523656783625, 0.381088763242587, 0.84821763751097, 0.510553280590102, 0.392026296816766, 0.592229992384091, 0.0951998953241855, 0.935264218831435, 0.0266276902984828, 0.594235675176606, 0.464057344011962, 0.522729218471795, 0.851427047746256, 0.124364921124652, 0.497395184123889, 0.949894226156175, 0.985215393593535, 0.196648222627118, 0.459533295826986, 0.662486824207008, 0.339795580599457, 0.907553537283093, 0.20275116735138, 0.019946081796661, 0.947368377819657, 0.774637680966407, 0.0241969616618007, 0.975378239061683, 0.958858958911151, 0.696549458196387, 0.73532759398222, 0.180152285145596, 0.767488027922809, 0.343676659977064, 0.435445247218013, 0.299161846050993, 0.248985140351579, 0.335878038080409, 0.268144787754864, 0.191027600783855, 0.034005434717983, 0.987843713257462, 0.800005949335173, 0.786398956552148, 0.376268550055102, 0.29064973955974, 0.76935594365932, 0.482855278067291, 0.48773535923101, 0.690200436627492, 0.969851274043322, 0.331725627649575, 0.333799781976268, 0.378115307539701, 0.0852395973633975, 0.590839970624074, 0.708808770403266, 0.647977532586083, 0.361773052951321, 0.947467828635126, 0.420819671824574, 0.751187635585666, 0.662229432258755, 0.775797713547945, 0.189702558331192, 0.214993652421981, 0.307810860918835, 0.814394601387903, 0.090209714602679, 0.171196923591197, 0.406157189980149, 0.205696649150923, 0.985900524072349, 0.696775345364586, 0.797266626963392, 0.544949859846383, 0.382226365851238, 0.283390154130757, 0.130625484511256, 0.965702779358253, 0.0881915763020515, 0.477424988290295, 0.947317285230383, 0.720382582396269, 0.42843632074073, 0.338347462238744, 0.613700705347583, 0.652796742739156, 0.507842858554795, 0.610279750078917, 0.159658169141039, 0.270344264339656, 0.13397484854795, 0.918471857439727, 0.929160196566954, 0.635510806925595, 0.170656314352527, 0.853746557142586, 0.956887910608202, 0.993407708359882, 0.0744277606718242, 0.486390518955886, 0.902719472534955, 0.507724171737209, 0.287784223211929, 0.745777606964111, 0.957658757921308, 0.198811845155433, 0.993768643587828, 0.892676381161436, 0.788876143284142, 0.712587109766901, 0.532617684220895, 0.291981202783063, 0.509067612234503, 0.773866353090852, 0.814828482922167, 0.566448303870857, 0.0634809567127377, 0.10337556921877, 0.27666057064198, 0.0785329001955688, 0.546709814341739, 0.393105684546754, 0.763675374677405, 0.160365695832297, 0.443380129523575, 0.837960895616561, 0.466975177172571, 0.338689843192697, 0.791645435150713, 0.169079039478675, 0.675454988377169, 0.975112377898768, 0.906052505131811, 0.477617010707036, 0.116861189948395, 0.640365410596132, 0.281903617782518, 0.0187481401953846, 0.938141646329314, 0.901878159027547, 0.661359721096233, 0.403403157601133, 0.7389045371674, 0.912085808813572, 0.283797955373302, 0.16165938670747, 0.0511467489413917, 0.505454917438328, 0.228351562051103, 0.142668584827334, 0.459028903394938, 0.341216093627736, 0.0423666436690837, 0.870234292233363, 0.0810602472629398, 0.761499754618853, 0.911109027685598, 0.770565738203004, 0.5070457062684, 0.0161868745926768, 0.667568264529109, 0.0135693300981075, 0.485727227060124, 0.12346867704764, 0.59017348755151, 0.871525567024946, 0.484375463565812, 0.920397447887808, 0.218572785612196, 0.802357716253027, 0.872094668913633, 0.276430501369759, 0.427874185144901, 0.617416493128985, 0.587696457048878, 0.551870324648917, 0.557024964131415, 0.751930515281856, 0.442513671470806, 0.823441645596176, 0.382196203339845, 0.564255010802299, 0.936024053720757, 0.323432239471003, 0.498935113660991, 0.118234819732606, 0.367213395191357, 0.796879748580977, 0.910054465057328, 0.716050766874105, 0.186144519597292, 0.52105862298049, 0.724362132139504, 0.564936319831759, 0.511641445336863, 0.728101488435641, 0.987818408757448, 0.518109863856807, 0.159865419380367, 0.74148246855475, 0.209912290796638, 0.721993474522606, 0.651955360313877, 0.587794766528532, 0.653149640653282, 0.474268056219444, 0.0670281811617315, 0.875122293597087, 0.844538746401668, 0.150132630486041, 0.561625748639926, 0.17210794868879, 0.700706083793193, 0.977902012178674, 0.351014475338161, 0.184257908724248, 0.316834892844781, 0.444335714215413, 0.530231991782784, 0.353868569247425, 0.00826845318078995, 0.656675655627623, 0.592926112702116, 0.809633676661178, 0.535552938468754, 0.284073724178597, 0.942641564179212, 0.386567050591111, 0.376742703374475, 0.447882670443505, 0.0638646888546646, 0.00618087546899915, 0.163446400780231, 0.631985178217292, 0.315025140997022, 0.732286646263674, 0.441260877763852, 0.0107825621962547, 0.634782533859834, 0.211565521545708, 0.509156787302345, 0.467435473576188, 0.937809493392706, 0.465590638341382, 0.0435755467042327, 0.0123572961892933, 0.179882112424821, 0.693633203860372, 0.608325648587197, 0.0790411829948425, 0.668177649378777, 0.900530697777867, 0.749251971486956, 0.26749463705346, 0.568183772731572, 0.963232085807249, 0.477401245152578, 0.679024447221309, 0.540779466740787, 0.18093074648641, 0.106794585706666, 0.948327814461663, 0.41636016801931, 0.582369825104252, 0.61759717296809, 0.513148660538718, 0.142035783734173, 0.0899850036948919, 0.613943622913212, 0.327450600452721, 0.529632689198479, 0.464640655787662, 0.334790169727057, 0.588361817644909, 0.417970288777724, 0.842360039474443, 0.0485526288393885, 0.670350446365774, 0.738777701277286, 0.618151533184573, 0.0318194003775716, 0.567022747127339, 0.0419600654859096, 0.378532435046509, 0.0150739345699549, 0.747771958587691, 0.705989859998226, 0.835282347397879, 0.178328054724261, 0.891039713053033, 0.774821633240208, 0.0773284879978746, 0.781065584858879, 0.385445648804307, 0.498060303041711, 0.0996416911948472, 0.752227542456239, 0.781906570773572, 0.778828218113631, 0.0452194116078317, 0.943698684684932, 0.713982340646908, 0.178900216473266, 0.871606213739142, 0.578695497009903, 0.663194889435545, 0.165322281653062, 0.537184630753472, 0.963698574109003, 0.369119519833475, 0.133231469895691, 0.328605210408568, 0.237311528995633, 0.132755285827443, 0.15870530414395, 0.325836643110961, 0.241155544994399, 0.835106596117839, 0.47487408365123, 0.0271863357629627, 0.787738534156233, 0.126608223188668, 0.484282916644588, 0.714774053776637, 0.465512123191729, 0.607501363847405, 0.217824660940096, 0.982783708721399, 0.57725323154591, 0.81584041682072, 0.357051741098985, 0.144364831503481, 0.0262512497138232, 0.409241791116074, 0.8083314539399, 0.311177365249023, 0.918827970977873, 0.645347734680399, 0.86923308763653, 0.51827756059356, 0.432039170991629, 0.846749583026394, 0.0486408618744463, 0.477585547603667, 0.196883853292093, 0.385982746724039, 0.0548241233918816, 0.419387371046469, 0.922276081517339, 0.665563164046034, 0.740310170454904, 0.432777700247243, 0.207278205547482, 0.0922720315866172, 0.802840916905552, 0.20616531255655, 0.535291479667649, 0.627356710610911, 0.982432310003787, 0.423445226158947, 0.212306978879496, 0.822790544014424, 0.418815913144499, 0.523124315077439, 0.138007411034778, 0.669121310114861, 0.894171869615093, 0.906509382883087, 0.946490505011752, 0.0756416202057153, 0.211675345664844, 0.359599119285122, 0.720532745821401, 0.811593460384756, 0.882577716838568, 0.403713824925944, 0.877383127808571, 0.283693644916639, 0.0829625432379544, 0.245037712622434, 0.607715327292681, 0.996052263770252, 0.953744772588834, 0.847680544247851, 0.448107725009322, 0.0544784460216761, 0.13796703889966, 0.384358690818772, 0.275465266313404, 0.363245291169733, 0.474431558279321, 0.267603539861739, 0.00810001604259014, 0.560508370399475, 0.124716287944466, 0.0597178800962865, 0.429863756755367, 0.755643857643008, 0.877224383177236, 0.0235673359129578, 0.984374420484528, 0.58708594343625, 0.358824983472005, 0.208523483481258, 0.622884845128283, 0.59162399196066, 0.676114350324497, 0.365566296037287, 0.121834203600883, 0.642142076743767, 0.117014496121556, 0.267002327600494, 0.0230049132369459, 0.780927070882171, 0.158872128929943, 0.963015584973618, 0.654179640347138, 0.648608357412741, 0.33610577788204, 0.650869886390865, 0.427160753170028, 0.0150227234698832, 0.488743406021968, 0.452819132013246, 0.600986144971102, 0.0625968785025179, 0.980248941341415, 0.438517449423671, 0.352537000086159, 0.740924185840413, 0.237253863597289, 0.851307639153674, 0.304173567797989, 0.475722135743126, 0.302097835810855, 0.272883016616106, 0.0945467478595674, 0.870235329959542, 0.0743445418775082, 0.403143124654889, 0.607056663371623, 0.97376810782589, 0.181924871634692, 0.195192408049479, 0.916052989428863, 0.818106508813798, 0.557930600829422, 0.826291869860142, 0.689399667317048, 0.323249742155895, 0.767605139641091, 0.85940212267451, 0.00101906177587807, 0.960645677754655, 0.46443037991412, 0.301900601712987, 0.609628821490332, 0.819510232424363, 0.394255306106061, 0.00841232179664075, 0.341141649987549, 0.638412500964478, 0.460150567349046, 0.288433650042862, 0.175102459732443, 0.0182740506716073, 0.82227867981419, 0.474351150682196, 0.808686204953119, 0.441384609322995, 0.278583735460415, 0.72582367900759, 0.037580557866022, 0.521430942462757, 0.0220515914261341, 0.1011481534224, 0.40235123434104, 0.041801555082202), actual = c(0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.330057704122737, 0.729311578907073, 0.510792092652991, 0.725068408763036, 0.754777594236657, 0.893207679735497, 0.783093282487243, 0.181564695434645, 0.775222759926692, 0.0197156155481935, 0.211597717599943, 0.854803936323151, 0.42047647992149, 0.843036873266101, 0.79574920097366, 0.402239270508289, 0.619223022367805, 0.837945051956922, 0.0899065041448921, 0.1204774514772, 0.686553715262562, 0.857188874157146, 0.0683495451230556, 0.764094315934926, 0.0564292811322957, 0.943699158728123, 0.0848500789143145, 0.541805095970631, 0.291150361998007, 0.715406357543543, 0.968694226350635, 0.327777759404853, 0.380558798555285, 0.433352330699563, 0.232887140708044, 0.71353337331675, 0.0496178064495325, 0.191977026872337, 0.74061689642258, 0.964090280467644, 0.727471817517653, 0.631123444065452, 0.841898065758869, 0.0624918963294476, 0.143308343598619, 0.251081966096535, 0.693752793828025, 0.0185899315401912, 0.557180475909263, 0.632085725897923, 0.678944887826219, 0.789038330782205, 0.979246493428946, 0.215195988304913, 0.444796750554815, 0.556571342051029, 0.606589748291299, 0.453080978244543, 0.0365875370334834, 0.711025996133685, 0.650296372128651, 0.60838800133206, 0.833076601382345, 0.547613228438422, 0.845206003403291, 0.330592697719112, 0.286340181017295, 0.664417843567207, 0.20883219782263, 0.940568304387853, 0.369224871043116, 0.084953400073573, 0.783875832566991, 0.335883195744827, 0.884713895618916, 0.365972510306165, 0.941879089223221, 0.815905786585063, 0.912050776882097, 0.0855883965268731, 0.429665397619829, 0.941614618292078, 0.0125936064869165, 0.419504404533654, 0.111848207656294, 0.167170372558758, 0.625689627369866, 0.143656614003703, 0.917117604753003, 0.170069268206134, 0.936426819534972, 0.749391294550151, 0.427252931054682, 0.249613302526996, 0.107287768041715, 0.598284466424957, 0.652906014351174, 0.337536850012839, 0.106230951612815, 0.470512378495187, 0.438250209204853, 0.612868272000924, 0.654572855914012, 0.69644873146899, 0.902853709412739, 0.231235288083553, 0.160141615895554, 0.355538899544626, 0.630017945775762, 0.306273267837241, 0.327449448872358, 0.265653174137697, 0.0783156640827656, 0.175576571607962, 0.337726780679077, 0.593103855848312, 0.423325426410884, 0.618307531112805, 0.15997572010383, 0.528382610064, 0.234961363486946, 0.481908258749172, 0.570512132719159, 0.413024924928322, 0.505251101450995, 0.312743459362537, 0.671705798478797, 0.051351779140532, 0.370604491326958, 0.285885192221031, 0.053351238835603, 0.546704340958968, 0.824874060228467, 0.674168206052855, 0.0497399740852416, 0.277927889721468, 0.950591743690893, 0.730555634945631, 0.784948628162965, 0.125149456318468, 0.110530447913334, 0.0582464297767729, 0.0269000728148967, 0.57119514956139, 0.329427517252043, 0.288441349053755, 0.359754791716114, 0.53112433408387, 0.646168014733121, 0.931004117941484, 0.866485723061487, 0.590407605981454, 0.217719252454117, 0.795666254358366, 0.740317060379311, 0.35810256912373, 0.0635889715049416, 0.0198493893258274, 0.499814320588484, 0.336417914601043, 0.46325828297995, 0.347142426297069, 0.30792221869342, 0.100392798194662, 0.922760728280991, 0.653804128291085, 0.238678352674469, 0.425995610887185, 0.272315677953884, 0.138616238720715, 0.783584929071367, 0.357843376463279, 0.677485981490463, 0.264331178041175, 0.943014926277101, 0.310126588912681, 0.300675431732088, 0.281337670050561, 0.799109728308395, 0.235161875374615, 0.520122274756432, 0.681681633926928, 0.0251439877320081, 0.00332944234833121, 0.831182589288801, 0.319875502726063, 0.816601180238649, 0.434805306838825, 0.553731770021841, 0.438514058012515, 0.195545091293752, 0.0535133755765855, 0.971193431410939, 0.424074118724093, 0.876235639676452, 0.156159204663709, 0.02288315189071, 0.701101433951408, 0.543967031640932, 0.780698469141498, 0.549240103922784, 0.80701725743711, 0.489412958500907, 0.753987843403593, 0.933682345319539, 0.591068963520229, 0.486327961552888, 0.874721253290772, 0.0610051341354847, 0.890286224894226, 0.120316844899207, 0.505409991368651, 0.117524623638019, 0.602455602958798, 0.738184981746599, 0.193001324543729, 0.384409070014954, 0.938510847277939, 0.648831222904846, 0.435491565149277, 0.742454027524218, 0.365432020276785, 0.203660548198968, 0.536122610094026, 0.385314833372831, 0.790510574588552, 0.837521638255566, 0.451578902080655, 0.146446831524372, 0.173180840210989, 0.600418791873381, 0.238440546439961, 0.492713797837496, 0.700142451794818, 0.936821497045457, 0.619512705830857, 0.94746253057383, 0.573599266586825, 0.298951605102047, 0.228349823970348, 0.472658462822437, 0.0117700907867402, 0.914624266326427, 0.911342387553304, 0.66152616054751, 0.801211496582255, 0.863928380887955, 0.0605109189637005, 0.888471347047016, 0.219897730508819, 0.740345110883936, 0.950119450455531, 0.294879670720547, 0.427714448655024, 0.6489096514415, 0.758024466456845, 0.325536935823038, 0.494803609326482, 0.605159452185035, 0.165093957213685, 0.859702231595293, 0.112693813862279, 0.409381592879072, 0.600578502984717, 0.842822869308293, 0.761321981204674, 0.619966094382107, 0.628772356081754, 0.800232905894518, 0.171968203037977, 0.727759754518047, 0.482289018807933, 0.509430071339011, 0.911041168263182, 0.122776974458247, 0.951781019335613, 0.395922740455717, 0.694873939733952, 0.376847212435678, 0.449596661375836, 0.470337982522324, 0.339676877483726, 0.10315245972015, 0.835453395266086, 0.0620507502462715, 0.190322282724082, 0.0437241462059319, 0.860732490196824, 0.610122695565224, 0.84098951308988, 0.546008165925741, 0.843674297444522, 0.821261778473854, 0.101074191508815, 0.45860594464466, 0.722464663442224, 0.720898469444364, 0.742170853773132, 0.628131044097245, 0.915381194325164, 0.885032387217507, 0.21151594328694, 0.52101371739991, 0.341056895675138, 0.467956540174782, 0.674738076748326, 0.268457878148183, 0.891103714704514, 0.458103453740478, 0.390200379304588, 0.903230241965503, 0.104773341910914, 0.35091680008918, 0.048621314810589, 0.724151507951319, 0.0984952403232455, 0.678122952347621, 0.66011810163036, 0.961052511120215, 0.71587070892565, 0.762914825696498, 0.0819927726406604, 0.258699357975274, 0.675922299502417, 0.585138511843979, 0.541450107702985, 0.271131973480806, 0.909253011457622, 0.63412378448993, 0.230508244363591, 0.299622604856268, 0.831198144936934, 0.0652764830738306, 0.363801765022799, 0.599132283125073, 0.0296100194100291, 0.81774139823392, 0.0623583674896508, 0.398077060701326, 0.301081691635773, 0.476329297758639, 0.982435496291146, 0.995276666479185, 0.635463654529303, 0.267726666294038, 0.363556284457445, 0.786251664161682, 0.198280234122649, 0.0963617933448404, 0.276120939524844, 0.880565059138462, 0.301687398459762, 0.0640166697558016, 0.308863073587418, 0.82988544809632, 0.118638807442039, 0.660179856931791, 0.471676732879132, 0.84318102686666, 0.944472608156502, 0.256488058250397, 0.324225985212252, 0.9129608431831, 0.12983038299717, 0.499923568451777, 0.691808725008741, 0.243411273462698, 0.217101831687614, 0.0329519922379404, 0.174969115993008, 0.20845955889672, 0.906784247374162, 0.639170175418258, 0.850916383555159, 0.803236462408677, 0.742217513499781, 0.667979455087334, 0.728629519697279, 0.530849867733195, 0.639218468917534, 0.216015659738332, 0.0115931264590472, 0.487907560076565, 0.432632491691038, 0.0396354498807341, 0.828881382476538, 0.437712979270145, 0.916121152229607, 0.93614765163511, 0.197974384296685, 0.0783294816501439, 0.314971412997693, 0.46792702213861, 0.110997091280296, 0.659762331284583, 0.349233741639182, 0.660089645534754, 0.607940625632182, 0.0292338605504483, 0.25213618692942, 0.943248512223363, 0.838400276377797, 0.718988733133301, 0.334697941085324, 0.422821899643168, 0.383063787594438, 0.520545807667077, 0.514310121536255, 0.691382624441758, 0.240948229795322, 0.0964525670278817, 0.676050957059488, 0.518027998507023, 0.401676748646423, 0.674619148951024, 0.874470928916708, 0.256723699858412, 0.430401963414624, 0.970359795494005, 0.807460210984573, 0.344615811016411, 0.727951947599649, 0.163494116393849, 0.13008771580644, 0.0940295499749482, 0.810406624339521, 0.904993337579072, 0.0819596112705767, 0.186534648295492, 0.952510713133961, 0.409513613441959, 0.31800883449614, 0.37404921441339, 0.65867456048727, 0.535763748222962, 0.55756958364509, 0.82455195649527, 0.970693449024111, 0.190359978238121, 0.41910247085616, 0.976554325781763, 0.952739477856085, 0.275548617355525, 0.712328071007505, 0.0936678671278059, 0.900408672168851, 0.630359938601032, 0.890029283007607, 0.739094311371446, 0.502027223585173, 0.784102610545233, 0.896224562078714, 0.706029518973082, 0.00658497726544738, 0.476187531836331, 0.678128664614633, 0.209276472451165, 0.43379679042846, 0.757473591016605, 0.626328242942691, 0.24065063544549, 0.629479738883674, 0.957552019506693, 0.402460426092148, 0.983293392928317, 0.802455852041021, 0.139572367304936, 0.97522384673357, 0.598460258450359, 0.968452326254919, 0.965243800543249, 0.614601846085861, 0.674971932312474, 0.917325054295361, 0.172710155369714, 0.886109074577689, 0.332652208628133, 0.33057179977186, 0.141431597759947, 0.595727553125471, 0.80895607243292, 0.732099737273529, 0.596663256641477, 0.030554513214156, 0.452251456212252, 0.588463450083509, 0.293888401705772, 0.976364499656484, 0.427526229294017, 0.818675988353789, 0.341680152807385, 0.973136842949316, 0.313468060456216, 0.206930983113125, 0.270310949534178, 0.777978860773146, 0.808127382304519, 0.843429871601984, 0.327004018006846, 0.845266045769677, 0.504695509327576, 0.264269894221798, 0.528701282339171, 0.256290724501014, 0.439191211713478, 0.983535398030654, 0.00261982041411102, 0.358801609370857, 0.862349095288664, 0.52005121530965, 0.847615759121254, 0.547880235360935, 0.166802154621109, 0.678506479132921, 0.0856798670720309, 0.948234789539129, 0.953657823381945, 0.156375881284475, 0.790414131712168, 0.632359266048297, 0.106572371907532, 0.161231436068192, 0.501687654526904, 0.509752585319802, 0.348619686672464, 0.829046547878534, 0.828657941194251, 0.498995565343648, 0.730128018418327, 0.737676292425022, 0.535947007825598, 0.212018533842638, 0.181638276902959, 0.655907125910744, 0.603258148767054, 0.551543383859098, 0.52216782537289, 0.789004976861179, 0.929197764256969, 0.979284485103562, 0.685776156373322, 0.957092782249674, 0.831539135891944, 0.352084443438798, 0.793590714223683, 0.495248758234084, 0.583320271456614, 0.729053685907274, 0.673242834629491, 0.584562242031097, 0.62716924957931, 0.770979670109227, 0.379283865215257, 0.193095307098702, 0.126341259805486, 0.117353465640917, 0.554472733289003, 0.0200597424991429, 0.572253664489836, 0.315189017914236, 0.231203005416319, 0.58988283504732, 0.1461051851511, 0.779922057176009, 0.0496600475162268, 0.398028040770441, 0.142569558229297, 0.0458231312222779, 0.828327244380489, 0.358761006267741, 0.54073612135835, 0.875729210209101, 0.179665223462507, 0.62231243099086, 0.306614047382027, 0.240284186089411, 0.699862326495349, 0.273389656329527, 0.72965145483613, 0.929802594473585, 0.958228669827804, 0.682860313681886, 0.867911629145965, 0.758049143711105, 0.896106890402734, 0.248903623549268, 0.723545809742063, 0.606472004437819, 0.0917066391557455, 0.758275555912405, 0.424451843136922, 0.379590429132804, 0.603684497764334, 0.721561504993588, 0.505421216133982, 0.450907375663519, 0.197703189915046, 0.905257569393143, 0.42324734525755, 0.0557101990561932, 0.0711195874027908, 0.22112000035122, 0.176123493118212, 0.840239677578211, 0.17601502710022, 0.210842156549916, 0.960983441444114, 0.27664085268043, 0.0921770355198532, 0.0666668822523206, 0.488678003661335, 0.909574567805976, 0.296502294018865, 0.770992995938286, 0.733279504813254, 0.686315467348322, 0.279685331508517, 0.58488344005309, 0.0256689202506095, 0.834966094465926, 0.0191219730768353, 0.419897966086864, 0.901575399097055, 0.574751785723493, 0.0799023157451302, 0.547619278077036, 0.0521457744762301, 0.98567022010684, 0.350442472379655, 0.187709144083783, 0.630633864784613, 0.797840610379353, 0.83338525891304, 0.420296164462343, 0.555451562162489, 0.644404769642279, 0.186534139094874, 0.0980658542830497, 0.916104468517005, 0.6354753409978, 0.171302637085319, 0.24184317747131, 0.201017648912966, 0.436394853517413, 0.92674716655165, 0.597726381616667, 0.620774062816054, 0.0437950482591987, 0.229321911931038, 0.471649211132899, 0.322507829172537, 0.593291280558333, 0.686796325957403, 0.11316214222461, 0.450695954961702, 0.712811261415482, 0.899375186068937, 0.303933978546411, 0.496992611093447, 0.77952217659913, 0.107828112784773, 0.984841795405373, 0.494751763064414, 0.410145868081599, 0.693892583949491, 0.822693368652835, 0.150593425147235, 0.485728629864752, 0.0839217668399215, 0.741033965023234, 0.236896121874452, 0.913344443775713, 0.0889273609500378, 0.962630430236459, 0.619519943604246, 0.982494363794103, 0.842063338961452, 0.021766884252429, 0.363672545878217, 0.971640542848036, 0.893182790372521, 0.318219518754631, 0.703622396569699, 0.711231584195048, 0.2306500650011, 0.844526654109359, 0.760037970496342, 0.85359823028557, 0.0245860849972814, 0.564384411554784, 0.0491375289857388, 0.252631842158735, 0.86394355725497, 0.838774153031409, 0.814230050193146, 0.195839480031282, 0.142549481242895, 0.949077371740714, 0.782934334361926, 0.914466488873586, 0.111418954329565, 0.612269634613767, 0.204768079100177, 0.511414432199672, 0.867947582853958, 0.60926823108457, 0.908047573175281, 0.668477257248014, 0.711314609274268, 0.950022696284577, 0.842934964457527, 0.734743291977793, 0.947429416235536, 0.889171577990055, 0.295059083029628, 0.0751779247075319, 0.405085396952927, 0.356444731820375, 0.704992482904345, 0.494124143384397, 0.258415828458965, 0.509162175003439, 0.0390077065676451, 0.199002478737384, 0.838492712937295, 0.809680094011128, 0.941873371135443, 0.575942856958136, 0.238537530181929, 0.196191709022969, 0.185122531605884, 0.256574444007128, 0.973820916842669, 0.562540218932554, 0.144867656519637, 0.0552579059731215, 0.615110820159316, 0.35601869574748, 0.869476899970323, 0.156316857319325, 0.548673656536266, 0.225790436845273, 0.554432794451714, 0.638806261820719, 0.817202369915321, 0.46579670207575, 0.933113213861361, 0.648737441282719, 0.550135362194851, 0.146861941553652, 0.0423086897935718, 0.0289948140271008, 0.647357150912285, 0.954880657605827, 0.534831339493394, 0.934252292150632, 0.861824658699334, 0.702627524267882, 0.98221934074536, 0.317652028752491, 0.563800428062677, 0.330046236515045, 0.13672478729859, 0.41657756222412, 0.511720176786184, 0.610318442573771, 0.293510216288269, 0.0550574874505401, 0.249319730559364, 0.990338939009234, 0.644321148982272, 0.0333396021742374, 0.324333297554404, 0.310182799817994, 0.141601920360699, 0.442538826027885, 0.1381180153694, 0.141002561897039, 0.701591827441007, 0.171915629412979, 0.127120318124071, 0.279703555395827, 0.108996706781909, 0.541469081072137, 0.513867403846234, 0.0172439387533814, 0.214246393181384, 0.131550067104399, 0.964366699801758, 0.595587597927079, 0.823848639382049, 0.256023785099387, 0.984788294881582, 0.0305229856166989, 0.0341897420585155, 0.080569313140586, 0.0309304588008672, 0.641106203896925, 0.590323691489175, 0.670765823451802, 0.323699656873941, 0.291246933396906, 0.199412069749087, 0.596263871062547, 0.903058124473318, 0.782179291825742, 0.417336388258263, 0.754080249927938, 0.228250463725999, 0.928116389084607, 0.822256865445524, 0.206937134731561, 0.665213861968368, 0.222060721600428, 0.131985170533881, 0.571720851585269, 0.55697157792747, 0.787254888331518, 0.550928112817928, 0.365919906413183, 0.0124686302151531, 0.819182829465717, 0.261717739049345, 0.219116047024727, 0.570345811080188, 0.506947074085474, 0.398002667119727, 0.0348863622639328, 0.465536762727425, 0.0919899444561452, 0.736834225943312, 0.151257372926921, 0.844729233998805, 0.158044582232833, 0.739828263176605, 0.337279246887192, 0.995684530818835, 0.302392415236682, 0.549657953204587, 0.305087980581447, 0.442878533853218, 0.539157556369901, 0.430572933750227, 0.538449191022664, 0.211558677721769, 0.106623706873506, 0.379580570850521, 0.252674009185284, 0.953918212326244, 0.13569157384336, 0.129128681262955, 0.737181433942169, 0.379766586236656, 0.495980222476646, 0.737263285554945, 0.457942426437512, 0.888130140257999, 0.100330493180081, 0.180878129554912, 0.127722747856751, 0.0652151529211551, 0.977890777634457, 0.237031441647559, 0.784623784245923, 0.0115185815375298, 0.0496467621996999, 0.785910340491682, 0.927826570579782, 0.22807564213872, 0.495790275046602, 0.829041157849133, 0.497671217890456, 0.186575008789077, 0.299921205732971, 0.7406329899095, 0.357603229116648, 0.708214428974316, 0.793146882206202, 0.801604021340609, 0.479774764506146, 0.665275739505887, 0.648424891056493, 0.7448005231563, 0.969338879687712, 0.991607749368995, 0.756101051811129, 0.889424381079152, 0.0521222432143986, 0.11221155943349, 0.315593778854236, 0.636695568449795, 0.150844365125522, 0.0751226041465998, 0.493903209920973, 0.218848725780845, 0.0732234446331859, 0.877371473936364, 0.539303172146901, 0.663915583398193, 0.984075471758842, 0.411262022797018, 0.374816952040419, 0.721762252273038, 0.503085697768256, 0.247535940259695, 0.526643596822396, 0.0529006985016167, 0.418738924199715, 0.909291933756322, 0.493939603678882, 0.15421276143752, 0.328525100834668, 0.027560779126361, 0.566076572285965, 0.43259992543608, 0.709800382843241, 0.0919805997982621, 0.86237603216432, 0.580693599069491, 0.550287865567952, 0.201290847733617, 0.667404550360516, 0.572840581415221, 0.366716209566221, 0.838318771217018, 0.220399201381952, 0.790596135659143, 0.0899479964282364, 0.133150826673955, 0.516468132846057, 0.496376252034679, 0.106088515603915, 0.826931049581617, 0.966501059010625, 0.110804476775229, 0.816291056573391, 0.318508847616613, 0.690684584900737, 0.384993945714086, 0.22766230488196, 0.18934093345888, 0.973252258030698, 0.0560817699879408, 0.279612331883982, 0.687914113979787, 0.640075684757903, 0.193325944477692, 0.159127674996853, 0.665972092188895, 0.50688134226948, 0.291848349617794, 0.643156173638999, 0.61804755916819, 0.337001154664904, 0.242125655757263, 0.602956532267854, 0.343095279531553, 0.203519185073674, 0.786158192669973, 0.0237812947016209, 0.698397226631641, 0.662496433360502, 0.368804092286155, 0.23056335397996, 0.248280008789152, 0.123777572531253, 0.285039108712226, 0.245262297336012, 0.202288263710216, 0.826130167813972, 0.315521599957719, 0.969992199679837, 0.59785807877779, 0.0135024562478065, 0.850366740254685, 0.633505403995514, 0.625870961230248, 0.858243460999802, 0.790071709081531, 0.307971936883405, 0.668259977595881, 0.465092658763751, 0.742487685056403, 0.69852711702697, 0.31144630885683, 0.0530815827660263, 0.485156965674832, 0.172871973831207, 0.844539985992014, 0.374642666894943, 0.858280830550939, 0.736116509884596, 0.233640559250489, 0.562281762482598, 0.0775574145372957, 0.298841423355043, 0.711736732628196, 0.301673606270924, 0.773102650418878, 0.0643060472793877), actual = c(0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"