Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.3361191        0.7880306 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5330280        0.5330280        0.5019934 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.600180788896978, 0.101375709753484, 0.568027124973014, 0.982444119174033, 0.840161327039823, 0.543505685171112, 0.524234662996605, 0.622137791244313, 0.0929966415278614, 0.475334313232452, 0.384226416703314, 0.000675740651786327, 0.835700884694234, 0.681452695745975, 0.227071300148964, 0.122764420462772, 0.566655994858593, 0.22278680303134, 0.30478915758431, 0.899886743398383, 0.343409456312656, 0.570237076375633, 0.752073065144941, 0.291710365563631, 0.81785358604975, 0.214617659570649, 0.739088621921837, 0.168246098095551, 0.594057763693854, 0.723412689520046, 0.846015973715112, 0.580298468703404, 0.384549483889714, 0.0163693169597536, 0.90412768162787, 0.870785182807595, 0.0575676949229091, 0.870105962036178, 0.712892617098987, 0.419053433695808, 0.179671765537933, 0.634683792479336, 0.11992086446844, 0.728743381798267, 0.0344879599288106, 0.0611749549861997, 0.228675055317581, 0.453236696077511, 0.00450033834204078, 0.0867369743064046, 0.198698104824871, 0.706179522909224, 0.905755435116589, 0.684974161908031, 0.110606579110026, 0.688329460797831, 0.881307628471404, 0.0233354547526687, 0.917538605630398, 0.242651200853288, 0.612688180059195, 0.489927826682106, 0.570696020731702, 0.3439946542494, 0.341301201377064, 0.224897872423753, 0.699624432250857, 0.803879718761891, 0.330026047537103, 0.85939337592572, 0.400724917650223, 0.053925346583128, 0.762982689077035, 0.437397267203778, 0.521019275533035, 0.0153756744693965, 0.903369619278237, 0.233186439611018, 0.673855194356292, 0.532412882428616, 0.855576992267743, 0.359165427042171, 0.481576144462451, 0.400896209524944, 0.426630540983751, 0.699015585472807, 0.384045522194356, 0.56267431541346, 0.915679179364815, 0.394530543591827, 0.339168842183426, 0.108239594614133, 0.147638051537797, 0.39233952597715, 0.466375959804282, 0.939271216280758, 0.217547018546611, 0.804573433939368, 0.866950331488624, 0.355181746650487, 0.772814265452325, 0.716890417039394, 0.24869603314437, 0.334446621127427, 0.0964242443442345, 0.0463366815820336, 0.387507343664765, 0.282664424041286, 0.181059881811962, 0.427814009599388, 0.10254976968281, 0.064106730511412, 0.693792837206274, 0.108493807259947, 0.915352504234761, 0.92493681772612, 0.00272008869796991, 0.855496757663786, 0.611961166141555, 0.568441556533799, 0.47078797291033, 0.448649728437886, 0.995407241629437, 0.570825035683811, 0.237420686054975, 0.42557857464999, 0.506170646753162, 0.503986253170297, 0.167399201309308, 0.194485252955928, 0.20399435935542, 0.201152979396284, 0.68342677410692, 0.464944918872789, 0.78730259812437, 0.202241723891348, 0.704898040741682, 0.809253158513457, 0.194574219640344, 0.0792637059930712, 0.838428541552275, 0.201645782450214, 0.778117794776335, 0.413264952367172, 0.437011273577809, 0.315241748001426, 0.0298690132331103, 0.448799944482744, 0.784320313716307, 0.327447604155168, 0.605455254437402, 0.475326608167961, 0.576630836585537, 0.0138408017810434, 0.130274198250845, 0.304749365430325, 0.0663444267120212, 0.689353612717241, 0.157499530352652, 0.53337103058584, 0.253846353152767, 0.364725636551157, 0.531686355592683, 0.674265270587057, 0.174020470352843, 0.350704497424886, 0.863421044778079, 0.240616721566767, 0.981106693390757, 0.719721825327724, 0.435905104735866, 0.166542997816578, 0.0251601741183549, 0.404439434874803, 0.344660660484806, 0.538479927228764, 0.403128071920946, 0.685134043684229, 0.367262480547652, 0.776479475200176, 0.585120434407145, 0.0145556007046252, 0.142982503399253, 0.340152026852593, 0.212735141627491, 0.892487894976512, 0.0144770566839725, 0.778637826209888, 0.412275133887306, 0.46264241496101, 0.914451665943488, 0.501030289335176, 0.349547612480819, 0.971054721390828, 0.932922193780541, 0.776303926482797, 0.883931141579524, 0.23088128794916, 0.635203141020611, 0.474120658356696, 0.575503234053031, 0.967627831734717, 0.160645340103656, 0.404439681442454, 0.906646365765482, 0.978119278559461, 0.626323674339801, 0.00189895997755229, 0.849530035397038, 0.691298586549237, 0.110573198646307, 0.881396537413821, 0.569104315713048, 0.377565340138972, 0.597855395404622, 0.34002797701396, 0.805231695994735, 0.736493254080415, 0.521980581572279, 0.268395940307528, 0.223227810347453, 0.322356844786555, 0.641408138908446, 0.954304477898404, 0.751420184038579, 0.652323688613251, 0.402430157409981, 0.971096683293581, 0.319813820999116, 0.522588779218495, 0.851634096819907, 0.676747574703768, 0.163419257150963, 0.408416577614844, 0.378386554541066, 0.135065394220874, 0.592267743311822, 0.58953611808829, 0.68431152170524, 0.461513268994167, 0.931055426830426, 0.741288146469742, 0.127454543486238, 0.0538397100754082, 0.0188976980280131, 0.377258583903313, 0.572431651409715, 0.748579340986907, 0.265522094909102, 0.690937368664891, 0.306556352647021, 0.178990215295926, 0.0990398542489856, 0.24745405302383, 0.638833972392604, 0.335194603772834, 0.733358773170039, 0.863110125996172, 0.0998690954875201, 0.460308331297711, 0.585436528082937, 0.833691380685195, 0.536647747969255, 0.511489687021822, 0.517586229369044, 0.155916897580028, 0.0557040881831199, 0.404535976471379, 0.00300476979464293, 0.620823871111497, 0.92611014097929, 0.771034004865214, 0.337659967830405, 0.0810840986669064, 0.344865496037528, 0.0258614921476692, 0.153552412055433, 0.855590484337881, 0.714872644748539, 0.554373906226829, 0.358876055805013, 0.368890227051452, 0.0647782643791288, 0.965400004992262, 0.656588486395776, 0.206049533095211, 0.681847051018849, 0.422671131324023, 0.195557786384597, 0.401442577596754, 0.386447452707216, 0.332749184686691, 0.124255519127473, 0.81022944720462, 0.396477654576302, 0.748031954746693, 0.803387719905004, 0.189859251258895, 0.250166235491633, 0.141758112004027, 0.484194954391569, 0.896129187662154, 0.490055983187631, 0.122934116749093, 0.83332711388357, 0.948504810454324, 0.985067081870511, 0.36265636398457, 0.295590528054163, 0.583459821529686, 0.0969504362437874, 0.125755739863962, 0.521921894513071, 0.832289554644376, 0.907329190988094, 0.909489429090172, 0.31673871865496, 0.0410884749144316, 0.300851131556556, 0.460625144885853, 0.0348395358305424, 0.140276625053957, 0.0740864488761872, 0.142900309991091, 0.680649595567957, 0.838783899787813, 0.815014558145776, 0.994441508548334, 0.580294780898839, 0.000832118326798081, 0.888224101392552, 0.383114906959236, 0.874460828723386, 0.838817153358832, 0.697263417532668, 0.593193768989295, 0.730080001987517, 0.616798119153827, 0.137316714506596, 0.426217140629888, 0.730582862626761, 0.420604368206114, 0.686586334137246, 0.923279372043908, 0.687925927806646, 0.505866115214303, 0.424634691560641, 0.850429932819679, 0.943295028991997, 0.785015935543925, 0.77513686963357, 0.358498907880858, 0.843967485241592, 0.845581667497754, 0.792059298604727, 0.892787834629416, 0.728470093104988, 0.494682396529242, 0.063495124457404, 0.800618697423488, 0.412291166372597, 0.77020983514376, 0.0607996585313231, 0.226634640712291, 0.279584662755951, 0.954709973651916, 0.671615499770269, 0.64827613113448, 0.608237351756543, 0.0290844447445124, 0.689058216754347, 0.452903625788167, 0.00333611411042511, 0.922753165476024, 0.966165945632383, 0.860268206335604, 0.717710891505703, 0.55195552483201, 0.312278332887217, 0.0398675946053118, 0.147097948938608, 0.592311476822942, 0.971394926775247, 0.705175741575658, 0.475799925858155, 0.282011257484555, 0.802851044340059, 0.401939031435177, 0.77666671318002, 0.120654044672847, 0.243242929922417, 0.599835985805839, 0.878776873694733, 0.310888005886227, 0.410153329605237, 0.485700772376731, 0.758474792353809, 0.201469501014799, 0.50923909409903, 0.355203729821369, 0.317525828490034, 0.376161502674222, 0.134836349170655, 0.820935984607786, 0.161681147292256, 0.439284514635801, 0.739209323888645, 0.994805368827656, 0.324077869998291, 0.174683967605233, 0.62676478177309, 0.837231965735555, 0.182760420953855, 0.863105740630999, 0.0969092804007232, 0.609553199028596, 0.0725736813619733, 0.122981437016279, 0.27656796737574, 0.293075614608824, 0.0119849389884621, 0.788184237666428, 0.883901215158403, 0.568540466018021, 0.287780125392601, 0.439668749226257, 0.217675998806953, 0.280840072780848, 0.389565160498023, 0.506377659039572, 0.0479291207157075, 0.063435748917982, 0.326066614827141, 0.183422362199053, 0.237149599473923, 0.730376589111984, 0.210592966759577, 0.994944935198873, 0.00979709485545754, 0.217795019736513, 0.164316083304584, 0.813850487582386, 0.114459698321298, 0.97353205550462, 0.908859654329717, 0.27253121114336, 0.904292234219611, 0.81141418335028, 0.43441015551798, 0.988476133672521, 0.525428829016164, 0.600629355991259, 0.216736699687317, 0.105471279006451, 0.940035347593948, 0.681995055172592, 0.688381868414581, 0.220376937184483, 0.294415714219213, 0.0284801893867552, 0.90541819203645, 0.180409288266674, 0.537767733214423, 0.640230917371809, 0.502334521850571, 0.541336524300277, 0.133473086636513, 0.824057685211301, 0.493907393654808, 0.514147442765534, 0.192113890079781, 0.764187491266057, 0.767724300501868, 0.717719034524634, 0.679409497184679, 0.133916253689677, 0.708903930149972, 0.968457127688453, 0.680734012275934, 0.455518880859017, 0.873048472218215, 0.845727083506063, 0.513157028472051, 0.734423895832151, 0.650873661972582, 0.607508125482127, 0.390717733418569, 0.77392862085253, 0.329772637691349, 0.473246730631217, 0.628671238431707, 0.444702069042251, 0.399729242781177, 0.987301479559392, 0.409834782825783, 0.375907454406843, 0.465344706783071, 0.22785701835528, 0.42706817551516, 0.626107015879825, 0.222971470560879, 0.961863285629079, 0.444881634088233, 0.779316542902961, 0.201747510815039, 0.970725210150704, 0.309168842621148, 0.0843089176341891, 0.0485388622619212, 0.407386564882472, 0.596048745792359, 0.452855298528448, 0.173786062980071, 0.489161235745996, 0.845558715984225, 0.0848128770012408, 0.312133582076058, 0.876697916304693, 0.617601234000176, 0.543081821873784, 0.891782996477559, 0.292328560492024, 0.700899870600551, 0.752983767306432, 0.740976421628147, 0.0275761908851564, 0.473400723654777, 0.386068085441366, 0.612525931326672, 0.437802181113511, 0.0109262093901634, 0.942497937008739, 0.818992101354524, 0.869370015570894, 0.455020970897749, 0.878316805697978, 0.0487058912403882, 0.880928054451942, 0.471192693803459, 0.0658574514091015, 0.977704080520198, 0.591654675547034, 0.0209376050624996, 0.869872972369194, 0.49684845819138, 0.985804572701454, 0.641544046113268, 0.128057790454477, 0.891628774348646, 0.752285486552864, 0.210489558055997, 0.708828237839043, 0.693373379297554, 0.616322559071705, 0.0640287268906832, 0.613480601459742, 0.499069451354444, 0.471906634513289, 0.953386861132458, 0.577451252611354, 0.464051600778475, 0.207788228290156, 0.187524303561077, 0.249747648835182, 0.197447512997314, 0.991572668077424, 0.217385708820075, 0.761017329990864, 0.791475105565041, 0.234399593435228, 0.598303595557809, 0.997459865408018, 0.911162771983072, 0.97866769740358, 0.178596144542098, 0.810324209043756, 0.211948827374727, 0.674472665879875, 0.450870471773669, 0.0558111942373216, 0.709990513976663, 0.629822155227885, 0.749865908408538, 0.192354671191424, 0.373971879016608, 0.540626564295962, 0.640904315281659, 0.103150730486959, 0.458379494491965, 0.285491243703291, 0.586084401700646, 0.0910869429353625, 0.871221216395497, 0.272943171672523, 0.199755168519914, 0.0665831505320966, 0.631953367497772, 0.820442223688588, 0.00689970748499036, 0.992510693147779, 0.986868088133633, 0.376480171224102, 0.452374558662996, 0.181161758955568, 0.0200043292716146, 0.0531305391341448, 0.436889636795968, 0.186231912812218, 0.132424957351759, 0.505856220377609, 0.187810904812068, 0.152540415525436, 0.634934759233147, 0.986540301237255, 0.584878836991265, 0.221772588556632, 0.629134929738939, 0.671344731934369, 0.608936953824013, 0.0950714158825576, 0.231273821322247, 0.29908494791016, 0.440979520091787, 0.764722361462191, 0.55861744703725, 0.421982697676867, 0.547977772308514, 0.641063038259745, 0.711862792028114, 0.0846677171066403, 0.619008030043915, 0.313443316379562, 0.308915534056723, 0.44193802587688, 0.285018464783207, 0.400731377769262, 0.477796604158357, 0.0344970310106874, 0.954460931243375, 0.545115632470697, 0.494444275274873, 0.439429075922817, 0.573863714467734, 0.543191857403144, 0.908736403798684, 0.792651203926653, 0.737096681492403, 0.403444019611925, 0.147073974832892, 0.571882218355313, 0.96343310456723, 0.579914590809494, 0.800018807174638, 0.973404251970351, 0.847129964735359, 0.274212188320234, 0.502633158117533, 0.82210653484799, 0.131844486109912, 0.523793121799827, 0.110266873147339, 0.749126535840333, 0.472046542912722, 0.490074746776372, 0.452337578870356, 0.700194693170488, 0.0522197301033884, 0.71797400387004, 0.44605879415758, 0.907302180537954, 0.561964203603566, 0.374315564753488, 0.0166789987124503, 0.515732788015157, 0.824437218252569, 0.0887435290496796, 0.27681314595975, 0.777775074588135, 0.851576865185052, 0.569300567498431, 0.811625978676602, 0.454521641135216, 0.407563755055889, 0.688178638461977, 0.525647557340562, 0.62187931779772, 0.747016882989556, 0.854851813288406, 0.615630617365241, 0.26943151652813, 0.973316486924887, 0.66156326700002, 0.303053299430758, 0.573717260733247, 0.204584764782339, 0.851004853378981, 0.241609313525259, 0.894716255366802, 0.372429852141067, 0.0249009211547673, 0.645673536928371, 0.219871565699577, 0.706496460363269, 0.0322753456421196, 0.725900091696531, 0.739443498896435, 0.783094192389399, 0.640998161165044, 0.0728668607771397, 0.0452548658940941, 0.301429291954264, 0.42964380257763, 0.245174996322021, 0.151544544845819, 0.782185143558308, 0.590791901340708, 0.561281308764592, 0.434263745322824, 0.931155699072406, 0.118756488664076, 0.347551460377872, 0.299748566700146, 0.114528699079528, 0.921701774233952, 0.0610661725513637, 0.661334558390081, 0.579248088411987, 0.832471141591668, 0.59855672204867, 0.900341136613861, 0.778591931564733, 0.438429469708353, 0.839069644687697, 0.486229597125202, 0.184402613667771, 0.977507200324908, 0.280435409629717, 0.452162296976894, 0.340219253441319, 0.191828106762841, 0.166138099506497, 0.0372241467703134, 0.892611034680158, 0.720256119500846, 0.720891050761566, 0.0307155556511134, 0.508646694477648, 0.546330830315128, 0.553445680066943, 0.603332063416019, 0.709363038884476, 0.21633890690282, 0.947442807722837, 0.957344713388011, 0.0198951216880232, 0.690010909223929, 0.339961963938549, 0.847976772114635, 0.749421873362735, 0.682375800563022, 0.598472494166344, 0.982622983399779, 0.859564993530512, 0.109177092090249, 0.228610642021522, 0.340481504565105, 0.115367066813633, 0.184576613362879, 0.179526153719053, 0.762744036037475, 0.573925734031945, 0.886927163461223, 0.247834423091263, 0.838042419636622, 0.0557159180752933, 0.765322794904932, 0.0976168061606586, 0.256402690894902, 0.356152122141793, 0.0898276190273464, 0.955019749701023, 0.780599099816754, 0.53377636265941, 0.314019519370049, 0.0123471294064075, 0.883808445185423, 0.154844261938706, 0.655935297487304, 0.311491080094129, 0.648894160753116, 0.888799725100398, 0.832750464323908, 0.0423214044421911, 0.102897954406217, 0.506484228884801, 0.505894679808989, 0.62611018656753, 0.713997765211388, 0.995502650737762, 0.411591849289834, 0.930795605527237, 0.415326837450266, 0.312669498380274, 0.18364419741556, 0.623887395719066, 0.958401840180159, 0.430237368913367, 0.786877120612189, 0.817591746570542, 0.625263662077487, 0.499725550180301, 0.166715873172507, 0.175435205455869, 0.834093004930764, 0.182460149750113, 0.37954381830059, 0.621852306183428, 0.698813434457406, 0.442871188512072, 0.751534343231469, 0.889624366769567, 0.768107756972313, 0.276042786659673, 0.104514944367111, 0.105443192413077, 0.0810476895421743, 0.379018546547741, 0.856340843485668, 0.64641984179616, 0.244920595083386, 0.448550417087972, 0.707952372729778, 0.631835193373263, 0.609389977995306, 0.576819537440315, 0.472899322630838, 0.611001394921914, 0.636891379021108, 0.178934468189254, 0.937571164453402, 0.408893219660968, 0.656975209480152, 0.313482102937996, 0.187956582056358, 0.474032989004627, 0.427012597676367, 0.738274419913068, 0.679407816613093, 0.586846680613235, 0.278560904087499, 0.0207974412478507, 0.405481119640172, 0.603553424589336, 0.812597124138847, 0.237345060100779, 0.32992095593363, 0.890940900892019, 0.260809535393491, 0.535777926910669, 0.614203569246456, 0.121021730359644, 0.678772727726027, 0.862025755923241, 0.930592463351786, 0.0329354328569025, 0.227535519516096, 0.358996057184413, 0.689579061232507, 0.884773865807801, 0.0537012834101915, 0.325636228779331, 0.108839914901182, 0.967923827236518, 0.630754549521953, 0.341979754390195, 0.383457577321678, 0.686051716795191, 0.977561520645395, 0.655525140697137, 0.451481817290187, 0.652138978475705, 0.792292829370126, 0.857174088712782, 0.173077587271109, 0.066293703392148, 0.877332272008061, 0.535479738609865, 0.218655344797298, 0.63580269832164, 0.831323612714186, 0.437265679473057, 0.549704199424013, 0.530847929418087, 0.386070981854573, 0.000695119146257639, 0.854601196013391, 0.887177608674392, 0.179173493757844, 0.979334368603304, 0.713598075788468, 0.92986736795865, 0.0132153693120927, 0.621718848589808, 0.164156166836619, 0.410133978817612, 0.871588092064485, 0.0470851163845509, 0.75551878567785, 0.197453283006325, 0.169111673021689, 0.200471303425729, 0.57471109344624, 0.440026788739488, 0.808460683561862, 0.732202304759994, 0.99372424185276, 0.261545156827196, 0.114892195677385, 0.370764972642064, 0.78709855931811, 0.996508635347709, 0.918106953380629, 0.34733377606608, 0.434727733256295, 0.713381044333801, 0.567906317068264, 0.401869103079662, 0.00190953654237092, 0.757983401417732, 0.580433142604306, 0.589582253247499, 0.592062776675448, 0.324328439077362, 0.916884004371241, 0.683367279591039, 0.715322741074488, 0.65757272625342, 0.43113260762766, 0.982328451704234, 0.283274119254202, 0.131316678831354, 0.28465856378898, 0.890700288582593, 0.385809286730364, 0.883329302072525, 0.245889521203935, 0.488031883956864, 0.634893944952637, 0.956386067438871, 0.49875134206377, 0.60161267593503, 0.485432566842064, 0.979694428388029, 0.877402145648375, 0.414803158026189, 0.375256897881627, 0.581329540349543, 0.89944381872192, 0.646286710398272, 0.76023412309587, 0.0462990710511804, 0.0814569280482829, 0.264781231759116, 0.362247240496799, 0.109675313346088, 0.259806894697249, 0.0641715659294277, 0.127247509779409, 0.304154643788934, 0.899576706578955, 0.410760702099651, 0.458347586216405, 0.576064497465268, 0.300222864374518, 0.89858250762336, 0.130136258900166, 0.644546281080693, 0.73429158818908, 0.0695285231340677, 0.881195807596669, 0.918342970311642, 0.243701182538643, 0.381429099943489, 0.951429332140833, 0.0971748982556164, 0.890920304926112, 0.902197989402339, 0.889550906838849, 0.0286689272616059, 0.861300756689161, 0.0347003964707255, 0.063062412198633, 0.530921334633604, 0.198248930508271, 0.733285683207214, 0.350707339821383, 0.247068285010755, 0.350112369749695, 0.585125620709732, 0.117394334869459), actual = c(1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.254368977155536, 0.855280021205544, 0.267260949127376, 0.174690302461386, 0.157154775457457, 0.612994964234531, 0.945005132816732, 0.478655664483085, 0.974781356984749, 0.956390085862949, 0.0805005650036037, 0.527918654028326, 0.24656209629029, 0.260496191680431, 0.512820873875171, 0.222153138602152, 0.871389298699796, 0.163693939568475, 0.714933902258053, 0.373005354311317, 0.374975987477228, 0.0626616484951228, 0.59650080697611, 0.632661739597097, 0.727095284266397, 0.622968074399978, 0.281183573650196, 0.494530241005123, 0.723073657834902, 0.00850510341115296, 0.208940169773996, 0.191020438214764, 0.708376110065728, 0.0414795824326575, 0.918057176982984, 0.804061975795776, 0.402730987640098, 0.698096332140267, 0.32501503941603, 0.19258862407878, 0.325782378902659, 0.25613689282909, 0.541453317040578, 0.253014776390046, 0.200334882596508, 0.068189162062481, 0.907341474667192, 0.539634451037273, 0.0273834750987589, 0.527684959582984, 0.0279532412532717, 0.268181996187195, 0.943815139820799, 0.607992874924093, 0.939938165713102, 0.243888494092971, 0.982447012094781, 0.811032021651044, 0.456183691974729, 0.633865475421771, 0.361617994727567, 0.133273921441287, 0.819009006489068, 0.425770949106663, 0.352918669814244, 0.615452081197873, 0.254719397053123, 0.818130428669974, 0.796337014297023, 0.934644155437127, 0.314964659744874, 0.02225780650042, 0.99544976092875, 0.843699993565679, 0.873448181664571, 0.971213306067511, 0.387345282128081, 0.860701406840235, 0.496787701966241, 0.782970956759527, 0.318714234977961, 0.556959379930049, 0.77888558479026, 0.13980301679112, 0.600615835748613, 0.200667403172702, 0.818875707453117, 0.918522279942408, 0.807577449362725, 0.654671684606001, 0.83817883534357, 0.0893657612614334, 0.594827563967556, 0.574752436485142, 0.10541608184576, 0.574844375951216, 0.347152324859053, 0.0323580440599471, 0.142716031521559, 0.279246120015159, 0.390034919138998, 0.452856556978077, 0.20891677425243, 0.923708068439737, 0.180263775866479, 0.0629372401162982, 0.715027092956007, 0.526430253405124, 0.624150512507185, 0.214130145031959, 0.265763351460919, 0.135405160253868, 0.726053978316486, 0.921080295694992, 0.0288613711018115, 0.0728064952418208, 0.366322559071705, 0.915307582821697, 0.123098241863772, 0.475064137484878, 0.884876218857244, 0.367592697963119, 0.49993875133805, 0.376129488926381, 0.547100368654355, 0.359342090785503, 0.232991850236431, 0.356929442379624, 0.222729022148997, 0.372594551881775, 0.0833735167980194, 0.769354722928256, 0.745322154369205, 0.552155817160383, 0.00218333769589663, 0.0150796815287322, 0.552630593068898, 0.0567927823867649, 0.800591499311849, 0.935963402502239, 0.00837658531963825, 0.317327553639188, 0.852932810550556, 0.236460409360006, 0.707415329525247, 0.430082360748202, 0.110749829560518, 0.895162022206932, 0.897829960100353, 0.041599728167057, 0.242611982161179, 0.991302353795618, 0.639468889683485, 0.540126361185685, 0.137031534221023, 0.293916126247495, 0.159443832701072, 0.233288994058967, 0.141054053092375, 0.487764086341485, 0.354348591761664, 0.800958091858774, 0.147346506128088, 0.260231607360765, 0.0666869168635458, 0.293355024652556, 0.859343447024003, 0.306404054397717, 0.0762094806414098, 0.15496046282351, 0.0945959885139018, 0.424164765980095, 0.903185187838972, 0.955834110500291, 0.530350582441315, 0.121137409703806, 0.9574911207892, 0.96114314487204, 0.742957992246374, 0.00598798971623182, 0.888114154106006, 0.333166729658842, 0.290323520079255, 0.305967234773561, 0.639893611194566, 0.672567278845236, 0.815262930002064, 0.722124967025593, 0.377341547282413, 0.780304853105918, 0.420402236981317, 0.197330329334363, 0.955742979887873, 0.458323258906603, 0.325243359897286, 0.764342027250677, 0.859677150379866, 0.736225432949141, 0.691902255173773, 0.884712647413835, 0.573056948836893, 0.754567119758576, 0.442756176693365, 0.343631320632994, 0.351002154638991, 0.045393404783681, 0.838993380777538, 0.46103830402717, 0.221841087331995, 0.290739994496107, 0.00160341733135283, 0.137041097041219, 0.695580189116299, 0.381069381022826, 0.776315139606595, 0.285384732531384, 0.517317510675639, 0.68860370782204, 0.301476853899658, 0.476129001704976, 0.920120077440515, 0.0760844661854208, 0.0750112219247967, 0.783149001887068, 0.0144704389385879, 0.699367181630805, 0.822736201807857, 0.998440494760871, 0.0651419153437018, 0.696191164664924, 0.938520358642563, 0.46568318339996, 0.158599088201299, 0.684404923347756, 0.068607660010457, 0.24901095754467, 0.00202987133525312, 0.570396983763203, 0.410057028289884, 0.0657127883750945, 0.657686978345737, 0.110090548405424, 0.845112984068692, 0.37432086118497, 0.299004689324647, 0.19714103685692, 0.789099020184949, 0.605656259460375, 0.511353627080098, 0.95189044252038, 0.134459096007049, 0.922552140429616, 0.233263797825202, 0.307825575349852, 0.481085993116722, 0.585632860660553, 0.80966318375431, 0.190989777212963, 0.995079642627388, 0.595696926582605, 0.822488478384912, 0.885991601971909, 0.360207873396575, 0.263187007047236, 0.0422522304579616, 0.938902139663696, 0.937828965252265, 0.627472712425515, 0.704910527216271, 0.632936416193843, 0.896292018005624, 0.357464549364522, 0.822426795260981, 0.2418841435574, 0.100597253767774, 0.712842961307615, 0.9640137180686, 0.474099661922082, 0.575528360670432, 0.370648306328803, 0.817297275643796, 0.757760929642245, 0.488950299099088, 0.79892484494485, 0.0471603036858141, 0.56052350345999, 0.767773392610252, 0.331924863392487, 0.801031966460869, 0.798731354530901, 0.442370500415564, 0.88961670268327, 0.505028504179791, 0.0595929147675633, 0.253977661021054, 0.486959268338978, 0.710082551231608, 0.927786764223129, 0.999313386855647, 0.341395187191665, 0.801449196413159, 0.503352943109348, 0.208972164895386, 0.184175932314247, 0.944661718560383, 0.579128399956971, 0.964844529516995, 0.850406222511083, 0.229301652871072, 0.0332832529675215, 0.722291067475453, 0.126755484146997, 0.767817440442741, 0.016508599743247, 0.0522487997077405, 0.831379583803937, 0.87658052961342, 0.361082601593807, 0.497896534390748, 0.46462699980475, 0.813635722268373, 0.835195318330079, 0.354890365852043, 0.505822505801916, 0.0173578434623778, 0.949429523432627, 0.604496834799647, 0.0861325839068741, 0.376973740989342, 0.846489445539191, 0.189921070588753, 0.584876242792234, 0.0173729627858847, 0.923182726604864, 0.94924020790495, 0.324472154024988, 0.723575231153518, 0.143952971091494, 0.346808027476072, 0.643123605055735, 0.859331749612466, 0.387050612131134, 0.250556951388717, 0.614139262121171, 0.424984552199021, 0.304731604875997, 0.469525189138949, 0.0769647110719234, 0.221182977082208, 0.375668137101457, 0.830572129227221, 0.168827628251165, 0.472046485403553, 0.405006200773641, 0.46262869075872, 0.799102173186839, 0.182597141945735, 0.30475913756527, 0.790534120984375, 0.308560022385791, 0.290558296721429, 0.131077101221308, 0.827916069654748, 0.103651310084388, 0.82849471620284, 0.602905295789242, 0.704422079026699, 0.0375062741804868, 0.0296971555799246, 0.0395651187282056, 0.655307739274576, 0.816009694244713, 0.0824088351801038, 0.16349736857228, 0.73371214652434, 0.487279316177592, 0.227553055156022, 0.507889779284596, 0.255461594322696, 0.657292190939188, 0.27438539895229, 0.320166372694075, 0.907535379752517, 0.785979103296995, 0.854500536108389, 0.780393041204661, 0.0402782049495727, 0.603989461902529, 0.759904732229188, 0.147131658159196, 0.361139221349731, 0.980048978002742, 0.732047143159434, 0.642404163256288, 0.479107120539993, 0.643042439362034, 0.688136482611299, 0.438573255203664, 0.96255375421606, 0.171438273508102, 0.20267443312332, 0.276164158480242, 0.825775382341817, 0.231261652661487, 0.350312271155417, 0.0346349969040602, 0.911972932051867, 0.897899225354195, 0.397381296614185, 0.608447194565088, 0.180558836320415, 0.902696249075234, 0.319549414562061, 0.841657559620216, 0.256721392972395, 0.0184691436588764, 0.510992488125339, 0.63247325271368, 0.157966228667647, 0.162752704694867, 0.473378932569176, 0.397109519457445, 0.692709595896304, 0.173724436666816, 0.074406998930499, 0.437105184188113, 0.844411330064759, 0.560033687856048, 0.116714939009398, 0.487674394622445, 0.115669370163232, 0.351637441199273, 0.642701098462567, 0.470333437435329, 0.253883569035679, 0.934146353974938, 0.750882085878402, 0.655138592002913, 0.0508814034983516, 0.374100996414199, 0.781757493503392, 0.328278213506564, 0.893763690488413, 0.626581777120009, 0.28875708184205, 0.291299613891169, 0.950792419491336, 0.915282171918079, 0.643838644493371, 0.621346112573519, 0.300062321126461, 0.199215282686055, 0.796350313350558, 0.987566214287654, 0.0602784543298185, 0.921992642339319, 0.149860698264092, 0.0203507207334042, 0.648270442150533, 0.0268847444094718, 0.533426213776693, 0.364019561093301, 0.424220487009734, 0.210797569714487, 0.00599274574778974, 0.858598286984488, 0.335404854966328, 0.0546025945805013, 0.870299492264166, 0.192335175117478, 0.743902730289847, 0.389116981765255, 0.129182443022728, 0.0539216070901603, 0.749817737611011, 0.186373000964522, 0.082096932688728, 0.376922106137499, 0.450045296223834, 0.967213447904214, 0.8366631611716, 0.133436958072707, 0.209290487458929, 0.167778305476531, 0.174450974212959, 0.116676240228117, 0.476670049130917, 0.244494297774509, 0.731140193296596, 0.790498366579413, 0.622918418375775, 0.790771405678242, 0.142593505559489, 0.2886586219538, 0.400245804339647, 0.741066838148981, 0.780744129559025, 0.377439137781039, 0.706650192849338, 0.997834003763273, 0.403769151307642, 0.43986856401898, 0.558905305573717, 0.0452151354402304, 0.843746825587004, 0.908601160161197, 0.678509109420702, 0.275804972741753, 0.178822130197659, 0.378851357847452, 0.43530869926326, 0.450165373506024, 0.51502376049757, 0.129591648001224, 0.520452013239264, 0.175487656379119, 0.101840065559372, 0.225597711047158, 0.581202311208472, 0.0229371942114085, 0.932942344574258, 0.038213221821934, 0.754206410842016, 0.248643790138885, 0.187027057167143, 0.155569283524528, 0.0781994196586311, 0.0253585483878851, 0.546370655531064, 0.374681122135371, 0.815690947463736, 0.557729298248887, 0.695398649200797, 0.207095227669924, 0.0730975100304931, 0.65053090476431, 0.86862455192022, 0.703291435725987, 0.167076056357473, 0.975502644199878, 0.00163132417947054, 0.212006399175152, 0.624972441932186, 0.0192007396835834, 0.0865296118427068, 0.874236933654174, 0.044040322303772, 0.643499569501728, 0.771658736281097, 0.115011675981805, 0.00865666614845395, 0.351520989090204, 0.907194474013522, 0.0382177021820098, 0.336972854100168, 0.0313695000950247, 0.0237458976916969, 0.385359843494371, 0.819415714591742, 0.296727741835639, 0.936378343263641, 0.204513368429616, 0.743835865752771, 0.128317918395624, 0.242966395569965, 0.203632212709635, 0.996780960820615, 0.70799327432178, 0.209677119506523, 0.988988367374986, 0.77313639014028, 0.0383775276131928, 0.757042646640912, 0.854285490931943, 0.731793292565271, 0.0300804474391043, 0.610852657118812, 0.531739369267598, 0.942261510528624, 0.327818428631872, 0.3840035952162, 0.0609707513358444, 0.605561584467068, 0.644599760416895, 0.962908883113414, 0.546333036152646, 0.351976633071899, 0.406542825745419, 0.932305197697133, 0.485030767740682, 0.301710202591494, 0.64380294107832, 0.099019736982882, 0.0550721872132272, 0.876251092646271, 0.932244277792051, 0.943387659499422, 0.584536153124645, 0.0926198130473495, 0.749507616041228, 0.383297686697915, 0.860071138013154, 0.0723450095392764, 0.180284570436925, 0.543973038205877, 0.329463227419183, 0.791397580411285, 0.850959989940748, 0.51243129838258, 0.31457541952841, 0.607592657674104, 0.129631240153685, 0.138740999391302, 0.390805772505701, 0.545487942872569, 0.805333906318992, 0.961994068929926, 0.881018431857228, 0.733541902154684, 0.98671471956186, 0.377769137965515, 0.284186053788289, 0.765274083474651, 0.615984530886635, 0.574420283781365, 0.494769115932286, 0.523078627651557, 0.777521225856617, 0.635304484982044, 0.968711717287078, 0.957443718798459, 0.596343853510916, 0.281570238061249, 0.605515373405069, 0.0336238055024296, 0.73363683396019, 0.315965497400612, 0.697618518490344, 0.788022479973733, 0.339388167951256, 0.605653629871085, 0.386883038794622, 0.272201880812645, 0.750361133133993, 0.673197347437963, 0.876394876977429, 0.737856056774035, 0.28616041620262, 0.121875991579145, 0.0247288786340505, 0.212367648724467, 0.779126453679055, 0.346401332411915, 0.452233731048182, 0.932653326075524, 0.352569847367704, 0.70913114096038, 0.353137663332745, 0.153172418707982, 0.430891146650538, 0.971603391459212, 0.206924831029028, 0.336631548358127, 0.167945283697918, 0.482565219281241, 0.703544704243541, 0.380817637080327, 0.598638959228992, 0.84592076158151, 0.966311702970415, 0.428553519537672, 0.237255567451939, 0.728819753741845, 0.876110623590648, 0.761554489377886, 0.770026294048876, 0.877080456819385, 0.847877298714593, 0.934699165867642, 0.353073133621365, 0.0893214319366962, 0.812762426910922, 0.790747708408162, 0.827059298520908, 0.900737893767655, 0.923829366685823, 0.0568807236850262, 0.155517336912453, 0.0333392377942801, 0.609436203492805, 0.531640275381505, 0.326812960440293, 0.516986226197332, 0.750641749007627, 0.329293062677607, 0.177796426229179, 0.830162171740085, 0.918932455824688, 0.855865995865315, 0.897032645065337, 0.822608713759109, 0.453977345023304, 0.361760493833572, 0.822182393632829, 0.518635156331584, 0.481587977614254, 0.806885875761509, 0.139742083614692, 0.898767687380314, 0.863553594099358, 0.698735575657338, 0.253150553442538, 0.560769689036533, 0.375450630206615, 0.306764642009512, 0.999027744634077, 0.295467609772459, 0.606670706067234, 0.61272388859652, 0.966568732867017, 0.142799462191761, 0.870511888293549, 0.542181115131825, 0.334102110937238, 0.747223634971306, 0.170195501763374, 0.273944519925863, 0.150559982983395, 0.0190008834470063, 0.37410580413416, 0.316508465679362, 0.328536094399169, 0.59051929670386, 0.942271164152771, 0.841633525211364, 0.632264094892889, 0.832491848850623, 0.812243128893897, 0.751242776634172, 0.831472292775288, 0.977789432741702, 0.602414477383718, 0.48917290661484, 0.931445031426847, 0.875287695787847, 0.660991790471599, 0.171181401005015, 0.446946744341403, 0.251400012755767, 0.780717631336302, 0.349667555652559, 0.791603139368817, 0.119587817462161, 0.943937332602218, 0.198645280906931, 0.998858593637124, 0.51799166877754, 0.655706481542438, 0.710938344476745, 0.5851023751311, 0.0211750098969787, 0.904631654033437, 0.805590957636014, 0.324032796779647, 0.161652182228863, 0.106874692486599, 0.367494214093313, 0.0372224990278482, 0.757420603185892, 0.651242516702041, 0.575572069967166, 0.26979183126241, 0.830221683485433, 0.0298004099167883, 0.724857443710789, 0.276792779332027, 0.790978632867336, 0.611005987273529, 0.924902726430446, 0.944257533177733, 0.109318761620671, 0.433724054368213, 0.829881703713909, 0.402030112454668, 0.512277999892831, 0.40786942653358, 0.236534855794162, 0.43273051828146, 0.195522175170481, 0.613006111700088, 0.600928076077253, 0.572966342791915, 0.349784455262125, 0.764326688135043, 0.846009126864374, 0.643198060337454, 0.979818626539782, 0.581642115255818, 0.102949926396832, 0.487755360780284, 0.252162465825677, 0.426214852603152, 0.168396850582212, 0.0774028915911913, 0.770986925577745, 0.57636178820394, 0.28548175515607, 0.603582732612267, 0.059368658810854, 0.56096542486921, 0.73081721062772, 0.965458985418081, 0.412586509017274, 0.448886548867449, 0.3479069233872, 0.053892589872703, 0.991117496741936, 0.153889793436974, 0.900631920900196, 0.324972974602133, 0.632573994109407, 0.541485072579235, 0.285526328487322, 0.703583904309198, 0.064739637542516, 0.183912746142596, 0.657805250957608, 0.128259773831815, 0.0159740310627967, 0.134950762381777, 0.0191863193176687, 0.487060271203518, 0.892896448960528, 0.724095769925043, 0.627499114954844, 0.104094890644774, 0.666737145278603, 0.637630482204258, 0.518742287298664, 0.199720836244524, 0.187367331469432, 0.206501059001312, 0.702120466856286, 0.348950121784583, 0.481398329138756, 0.645752699347213, 0.642686547711492, 0.893445269670337, 0.753990306751803, 0.753243561834097, 0.496107500977814, 0.93064762884751, 0.204107978846878, 0.318420404335484, 0.657944716280326, 0.041212550830096, 0.279498582473025, 0.907921073492616, 0.894320371095091, 0.51447995333001, 0.142006789101288, 0.403941502794623, 0.312998164910823, 0.901794349076226, 0.640014229808003, 0.942764173261821, 0.289541404228657, 0.0854744198732078, 0.411278425017372, 0.799636160256341, 0.0432769558392465, 0.573508626315743, 0.727944722864777, 0.113717042841017, 0.897374893305823, 0.537160771433264, 0.205989927751943, 0.729596683289856, 0.641772491158918, 0.374872403917834, 0.107225677464157, 0.424101427197456, 0.200518682133406, 0.871977686882019, 0.519073103554547, 0.492355147143826, 0.606788163539022, 0.147892985958606, 0.0436382836196572, 0.00884939101524651, 0.667678194353357, 0.116175055038184, 0.626274337759241, 0.650323482695967, 0.770118471933529, 0.0907645721454173, 0.731028067879379, 0.301264988258481, 0.0960841351188719, 0.631520258262753, 0.456366918515414, 0.077033405425027, 0.799489919329062, 0.938026991905645, 0.302504932740703, 0.302887884201482, 0.660579645307735, 0.842256345553324, 0.341340410057455, 0.575369836064056, 0.18206381611526, 0.692773613147438, 0.955543959047645, 0.616323693888262, 0.081584396539256, 0.68269184534438, 0.203376348363236, 0.908765910426155, 0.234559148084372, 0.772893171058968, 0.174075062619522, 0.890344025800005, 0.517919051228091, 0.808109196368605, 0.395438959822059, 0.182193991728127, 0.0908827830571681, 0.871547464746982, 0.402170965913683, 0.787941491929814, 0.333442306844518, 0.390789315104485, 0.134276537690312, 0.594447209266946, 0.883342547575012, 0.264181858394295, 0.581752728670835, 0.672230276279151, 0.645756951766089, 0.00385322305373847, 0.0253114050719887, 0.618025108007714, 0.260217744391412, 0.656505714403465, 0.897351702209562, 0.71919906232506, 0.803691002773121, 0.126049532555044, 0.644913312513381, 0.389621635433286, 0.102192539721727, 0.653243716573343, 0.0292055001482368, 0.200875998008996, 0.532378484029323, 0.531848990358412, 0.107558666262776, 0.151337931631133, 0.545927562750876, 0.98289635614492, 0.335050952853635, 0.336961045395583, 0.641689637908712, 0.769536320120096, 0.102669819258153, 0.124526781961322, 0.678993905428797, 0.892552841221914, 0.5664715282619, 0.110178666654974, 0.416402754141018, 0.955983001505956, 0.605721425963566, 0.20538229146041, 0.870655334554613, 0.678733604261652, 0.252232980215922, 0.519186913501471, 0.611230571987107, 0.270127282477915, 0.33409885992296, 0.556350244907662, 0.582435000222176, 0.802115858765319, 0.209782512392849, 0.635342756286263, 0.357743642525747, 0.186923143919557, 0.409377903677523, 0.24452922353521, 0.522749508731067, 0.316280485829338, 0.495051755569875, 0.940489602508023, 0.107524308841676, 0.938299427973107, 0.143762584542856, 0.0719032797496766, 0.477778551634401, 0.442549829138443), actual = c(1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"