Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4878058        0.4875633 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4880484        0.4880484        0.5254975 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.101180608849972, 0.671414569718763, 0.870256989495829, 0.0567559285555035, 0.2928558785934, 0.824257222237065, 0.0512709673494101, 0.823009005514905, 0.864495457615703, 0.37423361861147, 0.955332426121458, 0.210553010925651, 0.620544050587341, 0.996815979015082, 0.158620428759605, 0.419219468021765, 0.0514169044326991, 0.497703624190763, 0.211724150693044, 0.509924944257364, 0.724300079746172, 0.823843947378919, 0.0929936422035098, 0.339464216725901, 0.180919711245224, 0.694290795596316, 0.923594863386825, 0.51590721658431, 0.601096543716267, 0.838487888453528, 0.739722849801183, 0.382903779856861, 0.299030430614948, 0.334360261214897, 0.476922443136573, 0.367695688502863, 0.317531605949625, 0.176567309070379, 0.563663406064734, 0.751802462851629, 0.913511343300343, 0.289545162580907, 0.360440654680133, 0.867392909713089, 0.762350759468973, 0.148560686968267, 0.943281098967418, 0.127987972926348, 0.340358747169375, 0.425844993209466, 0.388060881989077, 0.327124355360866, 0.38718039682135, 0.609829651191831, 0.0628726852592081, 0.810786495683715, 0.440284213516861, 0.483992186374962, 0.981385589577258, 0.247793844202533, 0.198439594823867, 0.821792940841988, 0.448604315053672, 0.411535423016176, 0.0358462706208229, 0.424096927745268, 0.232777702854946, 0.820207335054874, 0.212839329149574, 0.81634297571145, 0.0504804623778909, 0.627334509743378, 0.02745466446504, 0.781590658938512, 0.919019328663126, 0.733195119770244, 0.181931608123705, 0.524404761847109, 0.485695211915299, 0.0613738854881376, 0.757367465412244, 0.0599742289632559, 0.477698506321758, 0.6503341041971, 0.530363711295649, 0.629002313828096, 0.295465339208022, 0.345938996644691, 0.17268579476513, 0.811437160242349, 0.598627094645053, 0.323826636886224, 0.430050299502909, 0.979803307913244, 0.729857138125226, 0.893287588609383, 0.501577717019245, 0.763060810510069, 0.887163634411991, 0.111481246771291, 0.605196511140093, 0.920405392767861, 0.80975422472693, 0.117502648383379, 0.672085754573345, 0.0342042264528573, 0.502037785947323, 0.558099291520193, 0.176772018196061, 0.521025622496381, 0.128755543380976, 0.436085688881576, 0.937855640891939, 0.759899705881253, 0.287709784228355, 0.145338676404208, 0.872945380397141, 0.391520138131455, 0.0218245091382414, 0.0944654962513596, 0.871228118427098, 0.484235403127968, 0.228000343078747, 0.421760192839429, 0.464075296884403, 0.310238096164539, 0.0683744547422975, 0.206343433121219, 0.00406964402645826, 0.306729186093435, 0.389463872183114, 0.582115584053099, 0.0356705433223397, 0.00290318531915545, 0.339296507881954, 0.944953902158886, 0.230230919783935, 0.69724417407997, 0.234490254661068, 0.756318788509816, 0.125946049811319, 0.555189707316458, 0.330470599699765, 0.898232367588207, 0.565189958782867, 0.909017828293145, 0.481276178266853, 0.605209404136986, 0.858750027837232, 0.247768601868302, 0.704924124991521, 0.579789507901296, 0.842313394881785, 0.313603609567508, 0.491945279063657, 0.300668883370236, 0.986164778005332, 0.894208201440051, 0.5016744278837, 0.156637365696952, 0.503652360057458, 0.10665891924873, 0.662708262214437, 0.925305535783991, 0.821308893617243, 0.154244761681184, 0.672278620069847, 0.89080029563047, 0.3708904706873, 0.870036372216418, 0.731526290532202, 0.119935448048636, 0.0210983096621931, 0.856993176974356, 0.167163315229118, 0.751557195093483, 0.146793718216941, 0.115322125144303, 0.160045953467488, 0.209953144658357, 0.0964325908571482, 0.886706898221746, 0.803979062708095, 0.98419023072347, 0.917072067968547, 0.572798756882548, 0.653584904037416, 0.229051731992513, 0.819908131379634, 0.0154100779909641, 0.501637886045501, 0.771879122825339, 0.400861907983199, 0.110884659690782, 0.486899518407881, 0.714651512447745, 0.8946186339017, 0.213432157412171, 0.245832082815468, 0.654063438763842, 0.376145805930719, 0.117178711108863, 0.458151427796111, 0.543578813783824, 0.543594832299277, 0.570419897558168, 0.900566564174369, 0.182577188126743, 0.180327635258436, 0.692525224061683, 0.04650933528319, 0.494878531666473, 0.763211083365604, 0.571815100731328, 0.75564564880915, 0.440582637209445, 0.669483192963526, 0.15284982859157, 0.715896651847288, 0.669967483263463, 0.297274055890739, 0.0668855200055987, 0.689159497153014, 0.503794837044552, 0.528937598224729, 0.62019880884327, 0.550889435457066, 0.273091436829418, 0.202808595029637, 0.000370073365047574, 0.334499115590006, 0.649530960712582, 0.618848510319367, 0.635374499717727, 0.274399458663538, 0.0833122814074159, 0.587695209076628, 0.707084069494158, 0.744333530077711, 0.703801516443491, 0.64552422077395, 0.750982424942777, 0.736486780690029, 0.234132080571726, 0.894772197352722, 0.561332853278145, 0.40051541174762, 0.67706160643138, 0.897589829517528, 0.831879744539037, 0.326230867998675, 0.679044590797275, 0.489517076406628, 0.210401382064447, 0.102777658030391, 0.341311934171245, 0.236070493934676, 0.95284225535579, 0.452267533866689, 0.925927954027429, 0.0567258677911013, 0.231602354673669, 0.196590665960684, 0.687523825326934, 0.658671746728942, 0.588500935817137, 0.250372818205506, 0.271572703262791, 0.8795469366014, 0.937787519302219, 0.161242138361558, 0.755282765254378, 0.190452959854156, 0.700482192914933, 0.578135471558198, 0.509111804887652, 0.475048940395936, 0.764380111126229, 0.340331635670736, 0.230849112616852, 0.538461125222966, 0.369644798338413, 0.101903550792485, 0.732335773762316, 0.758940547006205, 0.13512494857423, 0.732702418696135, 0.731776689877734, 0.981858935439959, 0.263667661231011, 0.348845312604681, 0.819673192920163, 0.934753389563411, 0.790208552032709, 0.475377735216171, 0.969449215335771, 0.778362067881972, 0.730813425499946, 0.146899006096646, 0.0348347891122103, 0.524956970475614, 0.72755089844577, 0.782400833209977, 0.90134802996181, 0.566735945641994, 0.0635894376318902, 0.867239885032177, 0.853934588842094, 0.952845291467384, 0.0111865024082363, 0.597357363440096, 0.341511955717579, 0.541501340689138, 0.915951754897833, 0.653670619940385, 0.0717501454055309, 0.30035375058651, 0.49879093747586, 0.367013089591637, 0.280828324845061, 0.479285432491452, 0.565615834435448, 0.408479602308944, 0.973075679736212, 0.896851178025827, 0.55791346449405, 0.210463233524933, 0.680249274242669, 0.467218345962465, 0.802150471834466, 0.101581948809326, 0.136429015779868, 0.796184701612219, 0.834699190920219, 0.299195146653801, 0.542450822191313, 0.313806885387748, 0.998571480391547, 0.230851684929803, 0.989563714247197, 0.016169392503798, 0.604210032848641, 0.110132083995268, 0.37522258120589, 0.162036729976535, 0.4190012908075, 0.414396051550284, 0.885387274902314, 0.0647737628314644, 0.567841352429241, 0.795719905057922, 0.583217170787975, 0.644652455579489, 0.482901057228446, 0.967775122961029, 0.840528800617903, 0.297801917651668, 0.731638359138742, 0.281517074676231, 0.30587192485109, 0.968746457481757, 0.915853606304154, 0.430020216153935, 0.346507514361292, 0.267575723584741, 0.848872146103531, 0.350769785465673, 0.150483524193987, 0.0903304293751717, 0.0762925809249282, 0.414960485184565, 0.681075837695971, 0.497091091237962, 0.756760277086869, 0.404290360631421, 0.0774803049862385, 0.824238686123863, 0.404436269076541, 0.687428832286969, 0.259991809027269, 0.653629339300096, 0.0110856138635427, 0.0790463008452207, 0.364212587708607, 0.0509567793924361, 0.0892842670436949, 0.925065427785739, 0.337805521441624, 0.310097132343799, 0.609208074631169, 0.70372044085525, 0.728347324999049, 0.205173627240583, 0.880493683274835, 0.783393899444491, 0.831336918054149, 0.165230897022411, 0.511304279556498, 0.918415860738605, 0.329339346848428, 0.975299380719662, 0.251539297169074, 0.391392514808103, 0.837106989230961, 0.843669516500086, 0.320445541059598, 0.044000648194924, 0.590080640744418, 0.092953227693215, 0.837875968776643, 0.807074346346781, 0.0446675892453641, 0.871501060435548, 0.609889916842803, 0.323612835491076, 0.747749550500885, 0.224953753408045, 0.484916065819561, 0.429109374992549, 0.450514468597248, 0.942118603503332, 0.972300798166543, 0.0923287917394191, 0.537519484525546, 0.0642362732905895, 0.688129000598565, 0.913436150643975, 0.363483093911782, 0.17425125092268, 0.958801450207829, 0.341315139783546, 0.500036590499803, 0.175721043022349, 0.693039394216612, 0.377572581637651, 0.0590453324839473, 0.347673570038751, 0.201630183262751, 0.0633366291876882, 0.29971268423833, 0.907168565317988, 0.0551311625167727, 0.878480015788227, 0.380683832336217, 0.337384863058105, 0.0922121626790613, 0.840503488201648, 0.871050553629175, 0.61041805613786, 0.623218168504536, 0.287999619496986, 0.608638270292431, 0.964123893994838, 0.324901503277943, 0.942371608922258, 0.959942406509072, 0.67998195393011, 0.887266078963876, 0.59766429127194, 0.655632853507996, 0.165264757815748, 0.118782512610778, 0.0784121153410524, 0.41257750056684, 0.417604554910213, 0.284731780644506, 0.628360575996339, 0.705603452166542, 0.777588390046731, 0.758165819570422, 0.969890957465395, 0.925012793624774, 0.805623272433877, 0.241334361489862, 0.995589294238016, 0.743154462892562, 0.523960650432855, 0.292818116489798, 0.601191306719556, 0.919304380426183, 0.608978933421895, 0.501084004994482, 0.111808829475194, 0.243325009476393, 0.0218740573618561, 0.84632036741823, 0.156362606910989, 0.126912102801725, 0.463166027795523, 0.0953802440781146, 0.653807942988351, 0.197440764866769, 0.414774173405021, 0.470681088045239, 0.737352065276355, 0.687752794474363, 0.801914181560278, 0.228803919861093, 0.604782762005925, 0.171537511982024, 0.803915647324175, 0.478649422992021, 0.300590996397659, 0.823786431457847, 0.724093235330656, 0.103113378630951, 0.0230647188145667, 0.555647438624874, 0.49006625614129, 0.636236947728321, 0.388922652462497, 0.637609283439815, 0.12971292831935, 0.896681352052838, 0.376479214290157, 0.512822564924136, 0.934166827704757, 0.883573208702728, 0.234772758325562, 0.52393056708388, 0.0257351924665272, 0.101463827071711, 0.237514834851027, 0.131983431056142, 0.401335250819102, 0.803451952990144, 0.519176852656528, 0.0492467859294266, 0.396574578946456, 0.262852579122409, 0.155982617754489, 0.293826735811308, 0.202453526901081, 0.494324539322406, 0.610246481839567, 0.119648580206558, 0.390698999399319, 0.663852575002238, 0.98043321701698, 0.473284443374723, 0.0934533171821386, 0.29995507071726, 0.123379248427227, 0.327332919696346, 0.921240834984928, 0.273736944654956, 0.0970697184093297, 0.0635445339139551, 0.607461826410145, 0.577253848314285, 0.200675880769268, 0.510319756576791, 0.928900920320302, 0.385633384110406, 0.266764931147918, 0.945502832299098, 0.730191622162238, 0.146788961719722, 0.315699036698788, 0.630130313104019, 0.11323995445855, 0.478642105124891, 0.308259860146791, 0.14655917417258, 0.0994016421027482, 0.765736455796286, 0.763618389843032, 0.0273333161603659, 0.542898786254227, 0.470342780230567, 0.905456327833235, 0.209502708399668, 0.241599950473756, 0.591351080685854, 0.0789193091914058, 0.50315685919486, 0.690390944015235, 0.246624853694811, 0.432126941857859, 0.985518134664744, 0.92463427875191, 0.74862481513992, 0.785492474911734, 0.87477998342365, 0.310346767306328, 0.409859088482335, 0.731600869679824, 0.060128946788609, 0.850941878510639, 0.0104572393465787, 0.72112457244657, 0.300868057878688, 0.957434235140681, 0.100261297775432, 0.222061211941764, 0.590158968232572, 0.210696004331112, 0.0746660940349102, 0.0230249920859933, 0.94181600981392, 0.65462026395835, 0.109212353825569, 0.24936346616596, 0.289377480279654, 0.500522778136656, 0.466857955558226, 0.860036787344143, 0.0425438499078155, 0.644893167773262, 0.507600185694173, 0.0383647105190903, 0.612910700729117, 0.897956281667575, 0.519250516779721, 0.823366533964872, 0.234548309119418, 0.378047962905839, 0.185041996650398, 0.946481751278043, 0.18068732554093, 0.594969851430506, 0.826095942640677, 0.875064371852204, 0.978162583196536, 0.426793911261484, 0.281539626885206, 0.658601314993575, 0.314747727010399, 0.841443137731403, 0.707824014592916, 0.164038782939315, 0.409674710128456, 0.799348591826856, 0.405888641951606, 0.0404802174307406, 0.646895881276578, 0.325926989782602, 0.657296754186973, 0.146286779781803, 0.0823337971232831, 0.163547819247469, 0.998630793299526, 0.864125669002533, 0.852428080514073, 0.224648582981899, 0.874057778390124, 0.248383663361892, 0.70065088942647, 0.800198503304273, 0.529616986401379, 0.124854059889913, 0.469416838139296, 0.101886705961078, 0.10041147028096, 0.0551532551180571, 0.769693228416145, 0.301068217027932, 0.882273737806827, 0.687835312448442, 0.0174835040234029, 0.707670888397843, 0.0786507714074105, 0.048787203617394, 0.0238579194992781, 0.503696794621646, 0.963189138798043, 0.268708673771471, 0.917235973989591, 0.51467354898341, 0.696635734057054, 0.965959270251915, 0.661463188473135, 0.161501554772258, 0.685996592510492, 0.538656518794596, 0.491924406727776, 0.157244132831693, 0.949579014675692, 0.734494739444926, 0.727608296088874, 0.933006167644635, 0.630317806499079, 0.98167311726138, 0.329550947295502, 0.953341407934204, 0.992992954328656, 0.264840556774288, 0.512415896635503, 0.212422813754529, 0.276687060249969, 0.857463268795982, 0.214240901637822, 0.608032965101302, 0.797347975661978, 0.644333811709657, 0.0538507003802806, 0.338934103259817, 0.614429448265582, 0.289966508978978, 0.0728736177552491, 0.620185270905495, 0.848124378360808, 0.836488542379811, 0.924219719599932, 0.927740406244993, 0.390534664737061, 0.545509147690609, 0.458778948290274, 0.299600874306634, 0.601318978937343, 0.405585432890803, 0.893438128521666, 0.693142895819619, 0.338659148896113, 0.310529239708558, 0.731378671713173, 0.986777560086921, 0.306074241641909, 0.831661787349731, 0.363241435727105, 0.740594734437764, 0.680053308140486, 0.988968200283125, 0.569185635307804, 0.334030565805733, 0.930589534109458, 0.468486059224233, 0.353937451262027, 0.902917541796342, 0.783253635978326, 0.299835045356303, 0.53062730957754, 0.138255312107503, 0.295441957889125, 0.661960831144825, 0.0529532297514379, 0.48752304376103, 0.970910605741665, 0.454523649998009, 0.642420186894014, 0.270774855278432, 0.0312609642278403, 0.0806733693461865, 0.937112323241308, 0.247725367778912, 0.307624025736004, 0.608841890003532, 0.259715972235426, 0.689964247401804, 0.260825571604073, 0.109522584360093, 0.796292807674035, 0.535296686226502, 0.497521139681339, 0.722913386067376, 0.0770694485399872, 0.273663021391258, 0.759091438958421, 0.0387647021561861, 0.284723948687315, 0.66557707823813, 0.169861629838124, 0.450496763456613, 0.0596021758392453, 0.0245159103069454, 0.936067182105035, 0.953631539596245, 0.772318228846416, 0.646934165153652, 0.431024948135018, 0.305278185056522, 0.315426856279373, 0.045767467468977, 0.0898580076172948, 0.8643785987515, 0.831888661719859, 0.589411929016933, 0.254560033092275, 0.492244634777308, 0.0888490264769644, 0.0106964451260865, 0.320792606333271, 0.59625410521403, 0.468830883735791, 0.19349113968201, 0.754481150303036, 0.130786262918264, 0.486555494833738, 0.2072758753784, 0.378157171886414, 0.486283223843202, 0.913636113516986, 0.596105324337259, 0.857909765094519, 0.152038683183491, 0.109780243132263, 0.253708644304425, 0.128035288536921, 0.112705489853397, 0.415900029474869, 0.109659772133455, 0.787489307345822, 0.226799599826336, 0.398568433243781, 0.936159079894423, 0.211967346258461, 0.14024149090983, 0.630563092650846, 0.753531741676852, 0.460990727646276, 0.0716858624946326, 0.0661940698046237, 0.738377938745543, 0.578062781365588, 0.243743556551635, 0.644420238677412, 0.844757876126096, 0.702306480379775, 0.74858790426515, 0.437406444689259, 0.979688436491415, 0.169057781575248, 0.626114668091759, 0.257280281744897, 0.918403985211626, 0.458030238514766, 0.650599464541301, 0.439669359009713, 0.792579049011692, 0.310069226194173, 0.918095191940665, 0.694693292025477, 0.245851433835924, 0.55662433966063, 0.587957967072725, 0.691361708100885, 0.954841562313959, 0.319629970006645, 0.199227921199054, 0.627560530556366, 0.307611086871475, 0.715529277920723, 0.736766140442342, 0.0586357789579779, 0.668798170285299, 0.837426562095061, 0.863010052824393, 0.182900896063074, 0.197759644594043, 0.390848032897338, 0.551130822161213, 0.404820777941495, 0.263409217819571, 0.371472516097128, 0.733999824849889, 0.482536191353574, 0.0114494778681546, 0.612188054481521, 0.228059070883319, 0.706730012549087, 0.246771284844726, 0.915052665397525, 0.518595156492665, 0.903422507690266, 0.512276329565793, 0.786420706892386, 0.546050626784563, 0.396986372536048, 0.379152184119448, 0.113167070318013, 0.260130519047379, 0.824146008584648, 0.955746247433126, 0.925548041239381, 0.517531650140882, 0.379648964619264, 0.132223740685731, 0.508325697854161, 0.456999938003719, 0.506206333171576, 0.688159070210531, 0.418904225807637, 0.774928799364716, 0.764153609285131, 0.210718061076477, 0.486926060635597, 0.13069679425098, 0.618446099339053, 0.0820258669555187, 0.181126807350665, 0.885047530988231, 0.787922630319372, 0.184824178460985, 0.743094517849386, 0.668886996107176, 0.526541810017079, 0.610286226728931, 0.834435118362308, 0.228121299296618, 0.902489945525303, 0.753486377187073, 0.436226821271703, 0.535930999787524, 0.817315036663786, 0.00321549200452864, 0.803948574699461, 0.344999543623999, 0.444075519917533, 0.333337686257437, 0.760834002867341, 0.9515041555278, 0.108842557063326, 0.964925500564277, 0.289370801532641, 0.997141365893185, 0.702855374198407, 0.344652483472601, 0.852194105507806, 0.546090890420601, 0.0269656837917864, 0.0757241889368743, 0.653139468049631, 0.894249747507274, 0.458569151116535, 0.951639514183626, 0.318874451564625, 0.513015005039051, 0.728359803091735, 0.421346247429028, 0.492799225961789, 0.119804936926812, 0.203795732231811, 0.41306315548718, 0.705796781927347, 0.636776616564021, 0.867331479210407, 0.510148346656933, 0.332748018205166, 0.0715037344489247, 0.972928954288363, 0.4077160556335, 0.138941128738225, 0.70003422931768, 0.914720692439005, 0.537693233694881, 0.7111642267555, 0.880934645421803, 0.82219875860028, 0.800354302395135, 0.128806413151324, 0.950829871464521, 0.808792643947527, 0.0178982024081051, 0.802136975573376, 0.312631587032229, 0.766309410799295, 0.00322666787542403, 0.495038245338947, 0.122833743924275, 0.588345479685813, 0.301253254525363, 0.0163728091865778, 0.271364243235439, 0.0269057520199567, 0.0872879193630069, 0.353998849866912, 0.145736939739436, 0.366128056310117, 0.893476316705346, 0.360336523037404, 0.2445626296103, 0.356096608331427, 0.282818239647895, 0.580199439777061, 0.730224982835352, 0.710048656910658, 0.157871725270525, 0.941820090636611, 0.677165326895192, 0.101312092039734, 0.717132158344612, 0.77771166828461, 0.482759410049766, 0.851719239959493, 0.36814971244894, 0.88453856902197, 0.481801269343123, 0.384205712703988, 0.300635169725865, 0.796858022455126, 0.516870677936822, 0.547247386770323, 0.4898874189239, 0.547957739792764, 0.339193708496168, 0.0536416685208678, 0.878766830544919, 0.00188020966015756, 0.859276997391135, 0.0727239062543958), actual = c(1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.073058458045125, 0.795214065117761, 0.998130646767095, 0.046205191873014, 0.893544262740761, 0.566625612555072, 0.705086248693988, 0.172675634268671, 0.398976846132427, 0.354444584343582, 0.476276798173785, 0.729236865183339, 0.406987709226087, 0.421677382895723, 0.193923067301512, 0.461762166116387, 0.106159495888278, 0.497415881371126, 0.152742886450142, 0.824883304536343, 0.230979162734002, 0.369985696626827, 0.310497279046103, 0.245713859098032, 0.636872782371938, 0.509448464261368, 0.855445458553731, 0.929847219726071, 0.742505116155371, 0.0135854056570679, 0.950914713321254, 0.504999736789614, 0.897087820572779, 0.407793659716845, 0.0542762998957187, 0.785307160113007, 0.883637657854706, 0.145740980049595, 0.834677097154781, 0.418190732365474, 0.922823562519625, 0.242581701138988, 0.593570036580786, 0.573457203572616, 0.906623090384528, 0.838262310018763, 0.0294293668121099, 0.726956342346966, 0.278417851310223, 0.559911839431152, 0.698550413129851, 0.276019970653579, 0.415615175850689, 0.657459071604535, 0.989591953111812, 0.0272878429386765, 0.912135890452191, 0.457179564516991, 0.346022054785863, 0.316045015351847, 0.445108931045979, 0.912222178652883, 0.714507922763005, 0.191580679267645, 0.350283741019666, 0.236119097331539, 0.0387137671932578, 0.690985163440928, 0.323127501411363, 0.156754838535562, 0.616913972888142, 0.621278862468898, 0.0882988080848008, 0.350239519029856, 0.890089416177943, 0.260680249659345, 0.548332971520722, 0.639646313618869, 0.0983961657620966, 0.129442319273949, 0.666824947344139, 0.916078337468207, 0.623949981760234, 0.697709762724116, 0.588309727376327, 0.0477101190481335, 0.683569781947881, 0.410251925000921, 0.517404491547495, 0.925271966261789, 0.63746550027281, 0.98555698688142, 0.713617821922526, 0.687542927218601, 0.806487467605621, 0.604701727395877, 0.96999005204998, 0.0249820495955646, 0.20111356372945, 0.392234385944903, 0.61636018496938, 0.698259891709313, 0.857210282469168, 0.258456045296043, 0.55278661288321, 0.639446165645495, 0.600184247363359, 0.33810258936137, 0.470432088244706, 0.0470976249780506, 0.368954758392647, 0.848814588272944, 0.488125806907192, 0.678039488149807, 0.996233351994306, 0.297377138631418, 0.128574545728043, 0.467282614437863, 0.110575314611197, 0.745532942004502, 0.526038258569315, 0.622347045922652, 0.87697274191305, 0.779520077398047, 0.0522781852632761, 0.421088548144326, 0.450530547881499, 0.349946149159223, 0.736611827509478, 0.229820065200329, 0.166395619278774, 0.485658473800868, 0.178589843213558, 0.558121057692915, 0.56933021498844, 0.51212758733891, 0.638442466501147, 0.729698333656415, 0.854243983980268, 0.167385162319988, 0.505693910643458, 0.464261221233755, 0.135605241637677, 0.652623928152025, 0.386379163013771, 0.764836841262877, 0.236690978286788, 0.319981963839382, 0.16033278265968, 0.761474995175377, 0.493054531514645, 0.0234294875990599, 0.651094449684024, 0.186664308188483, 0.00210154475644231, 0.0967778626363724, 0.585546807851642, 0.275384368607774, 0.47108868858777, 0.209596805972978, 0.145755106117576, 0.485876907827333, 0.582210156135261, 0.75012222933583, 0.149582725949585, 0.93172053527087, 0.610156112816185, 0.681267538806424, 0.240317926974967, 0.350503964349627, 0.285403857473284, 0.731298800557852, 0.895673174178228, 0.377550784265622, 0.122091572498903, 0.226043411763385, 0.0212762407027185, 0.113603533245623, 0.380817363038659, 0.313718333141878, 0.567925685783848, 0.631185566075146, 0.59580584615469, 0.943508501397446, 0.00897208042442799, 0.612167846877128, 0.179710590047762, 0.551152706844732, 0.848542313324288, 0.479756657732651, 0.0355325506534427, 0.320165506796911, 0.848146130563691, 0.378209969727322, 0.111031297361478, 0.517056811833754, 0.554263133555651, 0.660544615704566, 0.567273736698553, 0.10198155650869, 0.900206411723047, 0.84860704280436, 0.926175666041672, 0.0703122881241143, 0.112661846913397, 0.949371390743181, 0.891268943436444, 0.942686830414459, 0.922166630160064, 0.892731193220243, 0.520496263634413, 0.183655565371737, 0.0211055907420814, 0.478415599325672, 0.637456023134291, 0.703896299470216, 0.768058873480186, 0.000672940397635102, 0.119026449043304, 0.399087512400001, 0.51972777931951, 0.369936772156507, 0.292031033430248, 0.233282390981913, 0.649748236872256, 0.350934498943388, 0.881704338826239, 0.750108304899186, 0.0409994705114514, 0.849402162712067, 0.251110879238695, 0.626545020146295, 0.996339360484853, 0.287914433050901, 0.971999633824453, 0.924906684784219, 0.770018341718242, 0.182774198008701, 0.420519455801696, 0.873302034568042, 0.151072611100972, 0.530324475606903, 0.693795928964391, 0.596783681306988, 0.393919712398201, 0.144631878705695, 0.1397047014907, 0.322931970003992, 0.32700315467082, 0.534901806153357, 0.591828283388168, 0.248429872328416, 0.728017123416066, 0.152687777532265, 0.497920025605708, 0.482241570251063, 0.704189050942659, 0.782028709538281, 0.813212658045813, 0.222266130847856, 0.419941285159439, 0.556465534726158, 0.351832128828391, 0.283640164881945, 0.593086502281949, 0.0524930343963206, 0.712977164657786, 0.695101374294609, 0.672572479816154, 0.0336583666503429, 0.93449573777616, 0.728174210293218, 0.904157535871491, 0.564458044944331, 0.456221131840721, 0.100075045134872, 0.974161583930254, 0.424686864716932, 0.531502860831097, 0.780323370825499, 0.866005560848862, 0.472204379038885, 0.279500520555303, 0.603697214042768, 0.243607209995389, 0.254357101861387, 0.815170069225132, 0.0300545524805784, 0.138366080354899, 0.901811832096428, 0.481079068034887, 0.947396921925247, 0.286609392147511, 0.717429853510112, 0.150227187434211, 0.574138200376183, 0.889282346935943, 0.617045260965824, 0.908054494531825, 0.229662310564891, 0.427697239210829, 0.0899400163907558, 0.747501337435097, 0.774472663644701, 0.835822104942054, 0.817518571158871, 0.0738528158981353, 0.890836482634768, 0.765911556547508, 0.953080055071041, 0.919259101850912, 0.232480115955696, 0.0883427415974438, 0.905203906586394, 0.144511823542416, 0.994241138687357, 0.113179562846199, 0.784336481941864, 0.318347076652572, 0.895417127758265, 0.781852453015745, 0.0567677160724998, 0.428228692384437, 0.502378615783527, 0.043671685969457, 0.592309169238433, 0.239657099125907, 0.351396035635844, 0.141692547593266, 0.2093628088478, 0.582947754766792, 0.107128420146182, 0.750971384346485, 0.722515542060137, 0.418640539981425, 0.993324047420174, 0.979383943136781, 0.809212000807747, 0.37419964838773, 0.0158736016601324, 0.695568795083091, 0.279602106893435, 0.401486337650567, 0.297341603552923, 0.272746794158593, 0.897542686434463, 0.876854725647718, 0.26265955530107, 0.369385290890932, 0.687959901988506, 0.103149473899975, 0.612783548422158, 0.319194871466607, 0.180525549221784, 0.1652383194305, 0.126994589809328, 0.162559900898486, 0.308900590054691, 0.885765214916319, 0.198517069686204, 0.832513075321913, 0.0130742264445871, 0.00845246249809861, 0.958609311608598, 0.0461868140846491, 0.597043354529887, 0.542964555555955, 0.235445105470717, 0.0374912298284471, 0.317364115966484, 0.158420398365706, 0.831301336642355, 0.0132707171142101, 0.60284189786762, 0.267320258542895, 0.768315049819648, 0.188345037866384, 0.614242031937465, 0.772544158855453, 0.549925632541999, 0.819525935919955, 0.989098979858682, 0.675969353877008, 0.0329393630381674, 0.534752289066091, 0.208478047512472, 0.400416020071134, 0.352838999591768, 0.56680749845691, 0.220956728328019, 0.0403186539188027, 0.505167971830815, 0.150379888014868, 0.0254683664534241, 0.685729638440534, 0.62216999894008, 0.623889193870127, 0.334320592926815, 0.642703815363348, 0.834044689079747, 0.994821497704834, 0.0614833531435579, 0.193042060127482, 0.789305647136644, 0.0305843672249466, 0.390377182746306, 0.278967125806957, 0.443439274094999, 0.0508677989710122, 0.0724888197146356, 0.152501292759553, 0.639783383812755, 0.846365347737446, 0.531888242810965, 0.820338748162612, 0.523178273113444, 0.0635555635672063, 0.139255269197747, 0.440825339406729, 0.692708264803514, 0.768717526923865, 0.394667179556563, 0.355132837314159, 0.0504038336221129, 0.240028709173203, 0.231789097189903, 0.617460512556136, 0.880804107291624, 0.217260011238977, 0.553588671376929, 0.498399421107024, 0.474817148409784, 0.0311509433668107, 0.64246755791828, 0.0707919157575816, 0.592905796598643, 0.337478206027299, 0.613026670413092, 0.478377912193537, 0.963455914752558, 0.078589127631858, 0.952096693450585, 0.975706503260881, 0.358276156475767, 0.460850590839982, 0.0622851878870279, 0.591208792524412, 0.106424840399995, 0.777531206142157, 0.540941495914012, 0.558296244591475, 0.0885640466585755, 0.779275095788762, 0.304176648845896, 0.57021050946787, 0.942549948347732, 0.14226639084518, 0.819752168375999, 0.0233036298304796, 0.797994148219004, 0.328826491720974, 0.69936214457266, 0.551229211501777, 0.934889493975788, 0.737058602739125, 0.944547366118059, 0.937841046368703, 0.864418976241723, 0.13305059960112, 0.666742453118786, 0.8939436187502, 0.777020719600841, 0.955282781971619, 0.857580995419994, 0.856113621965051, 0.807976945536211, 0.991378508042544, 0.86240098788403, 0.380146025912836, 0.689345393097028, 0.462412285385653, 0.777409762842581, 0.644161917502061, 0.333649680949748, 0.485836032778025, 0.767654044087976, 0.684046312933788, 0.314983120188117, 0.765408081002533, 0.809675226453692, 0.931461109546944, 0.315006234915927, 0.603585271863267, 0.823332633357495, 0.132318572606891, 0.31174451415427, 0.672295589698479, 0.92511799512431, 0.305403689853847, 0.114062990760431, 0.109146922128275, 0.60013683559373, 0.026492609642446, 0.328627176815644, 0.538677557604387, 0.674356810050085, 0.839617077261209, 0.364421909209341, 0.471120136557147, 0.629702299367636, 0.532623705919832, 0.220292051322758, 0.808395173866302, 0.814294615527615, 0.246391127817333, 0.244252485921606, 0.363913896493614, 0.791983666596934, 0.788999715354294, 0.571174876531586, 0.600382204866037, 0.551012207521126, 0.634634729474783, 0.938115991652012, 0.365030799759552, 0.272708925418556, 0.107309273211285, 0.731707359664142, 0.0367221024353057, 0.460656769340858, 0.0616879847366363, 0.777660630177706, 0.217782014748082, 0.786097457632422, 0.274474416160956, 0.242353135487065, 0.87201118352823, 0.169756176415831, 0.612473865272477, 0.495662017492577, 0.379494674503803, 0.350038899807259, 0.72779223928228, 0.96647380804643, 0.995144409127533, 0.105451809242368, 0.278832319192588, 0.0581297320313752, 0.869959869189188, 0.315541045274585, 0.858433150453493, 0.503924039192498, 0.102674246067181, 0.847638963023201, 0.579816391458735, 0.830124858068302, 0.19238655990921, 0.185543318977579, 0.974730195011944, 0.172747601289302, 0.387830417137593, 0.292110319016501, 0.0985213690437376, 0.765325586078689, 0.0872175339609385, 0.103052849881351, 0.0727087750565261, 0.351464090868831, 0.52820832352154, 0.0387637841049582, 0.79173540067859, 0.119232551660389, 0.616915317019448, 0.0370411893818527, 0.154126489302143, 0.419777463888749, 0.130923321237788, 0.200012423098087, 0.566364150261506, 0.0266375083010644, 0.497004331322387, 0.988248711219057, 0.155148168792948, 0.526305947452784, 0.932671329705045, 0.513822097796947, 0.113561464240775, 0.820599491707981, 0.0762702904175967, 0.0835590842179954, 0.277129673864692, 0.381464518373832, 0.735798645298928, 0.0845078346319497, 0.0388780585490167, 0.852336699608713, 0.945963982027024, 0.495894802268595, 0.580857424065471, 0.278567815665156, 0.359896950190887, 0.47077182517387, 0.672806233167648, 0.190780520904809, 0.153533447766677, 0.81676276633516, 0.843396130949259, 0.95854471065104, 0.977828710107133, 0.339166866615415, 0.971078650327399, 0.789778261911124, 0.0253303567878902, 0.804538538679481, 0.259425534168258, 0.121779198758304, 0.437023553764448, 0.383795270230621, 0.256781147094443, 0.805044820997864, 0.987800223985687, 0.790200646966696, 0.397343170130625, 0.165400456171483, 0.416971006430686, 0.402227231068537, 0.0753754565957934, 0.837708252016455, 0.0688855042681098, 0.198315352201462, 0.827201611362398, 0.250268779927865, 0.920605121878907, 0.140362157020718, 0.669701495906338, 0.48500998923555, 0.226106418063864, 0.808447104413062, 0.154556427616626, 0.274137549102306, 0.876656076870859, 0.279955903999507, 0.822763644857332, 0.885601912625134, 0.0943696477916092, 0.786263975780457, 0.756343126064166, 0.977254208410159, 0.237797261215746, 0.521067421883345, 0.431172606069595, 0.650776217225939, 0.947281835367903, 0.598270237911493, 0.0472681315150112, 0.404873134568334, 0.935871860710904, 0.39231171598658, 0.767866797279567, 0.813321561319754, 0.605095880571753, 0.767487041419372, 0.0166804366745055, 0.772454749094322, 0.834175723139197, 0.442679775180295, 0.349536954425275, 0.609635662287474, 0.812478991458192, 0.167878886917606, 0.745230742264539, 0.57122554210946, 0.695636255666614, 0.835385932354257, 0.091680308803916, 0.419955868739635, 0.573800627375022, 0.619546247413382, 0.893544732360169, 0.152258241316304, 0.0956270527094603, 0.524135659448802, 0.716670512920246, 0.110579555854201, 0.321121555753052, 0.593502667499706, 0.0393192912451923, 0.271952357841656, 0.928488861536607, 0.254875178914517, 0.361293436726555, 0.390391249908134, 0.8765996505972, 0.179060069378465, 0.890662077115849, 0.557046146597713, 0.39347775792703, 0.476177506148815, 0.560758477309719, 0.853042436996475, 0.0304210523609072, 0.456377546302974, 0.402287331176922, 0.543815845623612, 0.455649837153032, 0.607746853725985, 0.136015537427738, 0.455338674364612, 0.966804405674338, 0.730924260104075, 0.651925018522888, 0.0944039351306856, 0.2322798371315, 0.494382414733991, 0.0101063188631088, 0.346260359045118, 0.783839406445622, 0.802739782258868, 0.0401346501894295, 0.737642551306635, 0.849921832093969, 0.949204053962603, 0.28979663271457, 0.524196509970352, 0.919167748186737, 0.822891645599157, 0.766523320693523, 0.590776235563681, 0.851544338045642, 0.905381717486307, 0.0908089207950979, 0.796539625851437, 0.168042135890573, 0.224194626556709, 0.305625139269978, 0.378219413338229, 0.0904642310924828, 0.0348636012058705, 0.927035980625078, 0.763562186388299, 0.40706610144116, 0.210909852292389, 0.394819889916107, 0.580222714459524, 0.693588973255828, 0.0114463260397315, 0.368448911001906, 0.15829141298309, 0.105444764019921, 0.0110559822060168, 0.92049782932736, 0.379549667239189, 0.116208731429651, 0.612142141209915, 0.566940737189725, 0.00558506906963885, 0.563582041533664, 0.261910875560716, 0.350038231583312, 0.404912758385763, 0.0604899129830301, 0.252613951684907, 0.136589126894251, 0.258669339353219, 0.742900602985173, 0.921467391308397, 0.613142166752368, 0.781971732387319, 0.807171312626451, 0.647832259302959, 0.516400545369834, 0.0891458569094539, 0.262453723698854, 0.579080919735134, 0.879562191432342, 0.0139629044570029, 0.447632645256817, 0.407217907952145, 0.151130649494007, 0.314312178874388, 0.726869793608785, 0.0880050733685493, 0.542641005013138, 0.461082942085341, 0.735841631656513, 0.421160724945366, 0.340288165723905, 0.867975900415331, 0.152520244708285, 0.579204032430425, 0.322953204857185, 0.996698271017522, 0.355316187022254, 0.307766271755099, 0.960661836666986, 0.639242619974539, 0.0644336990080774, 0.626022164011374, 0.123100125929341, 0.709066255018115, 0.304247218882665, 0.377306633628905, 0.510412025963888, 0.485952673712745, 0.583363013574854, 0.945375200593844, 0.159573850920424, 0.691046088002622, 0.349502874771133, 0.317914849612862, 0.457833085441962, 0.963726131012663, 0.803133473964408, 0.46356881223619, 0.698674158658832, 0.873634721152484, 0.91717564011924, 0.737926865462214, 0.213996330043301, 0.807604887057096, 0.656275786459446, 0.0674534954596311, 0.0338471685536206, 0.858862848952413, 0.539085116237402, 0.954161166446283, 0.970994924427941, 0.759156569838524, 0.557060915511101, 0.792346892179921, 0.657012272858992, 0.836546040838584, 0.21223068726249, 0.976092780474573, 0.777401004452258, 0.564965688856319, 0.44622384570539, 0.00443977094255388, 0.960310378810391, 0.274295678827912, 0.438260044204071, 0.141793684568256, 0.309560050955042, 0.962335458723828, 0.925514356000349, 0.321157463360578, 0.709537284681574, 0.201097301673144, 0.936194317648187, 0.558134943945333, 0.388905055588111, 0.0738405752927065, 0.801528509939089, 0.360792764928192, 0.933852795045823, 0.467359349597245, 0.610589594813064, 0.157967553474009, 0.867388079408556, 0.78930118586868, 0.721379586262628, 0.757456608815119, 0.785406962037086, 0.436697498429567, 0.448863844852895, 0.99692518170923, 0.929708728333935, 0.495822707889602, 0.213030183222145, 0.0214815957006067, 0.478413061937317, 0.100561226252466, 0.890538181411102, 0.121200567344204, 0.824816769221798, 0.0649270184803754, 0.316449948120862, 0.543511593015864, 0.987130580469966, 0.614076283527538, 0.130976342828944, 0.662821947131306, 0.926557675004005, 0.607810628134757, 0.766808128915727, 0.523136201314628, 0.295307329390198, 0.731882181717083, 0.340431754011661, 0.836529608815908, 0.979415036970749, 0.870139508973807, 0.0713005647994578, 0.258355148369446, 0.770198848564178, 0.00979206594638526, 0.686554819811136, 0.629920759703964, 0.271311923163012, 0.484467072412372, 0.740308246109635, 0.683102554641664, 0.707724755397066, 0.661336044082418, 0.216073540737852, 0.777229167288169, 0.0020971829071641, 0.996105971047655, 0.920290300389752, 0.75305103790015, 0.600388902472332, 0.13372362148948, 0.515756766078994, 0.105948262149468, 0.725400853669271, 0.432382245315239, 0.777327069081366, 0.729643810773268, 0.473250386305153, 0.489381974795833, 0.174357475014403, 0.076895943377167, 0.844071545638144, 0.802881689509377, 0.737470969790593, 0.920167345320806, 0.142778239911422, 0.775516794994473, 0.370906998869032, 0.24556724447757, 0.500186364166439, 0.122803853126243, 0.37886799313128, 0.0198461615946144, 0.988652609055862, 0.42894863919355, 0.214379074750468, 0.0405819821171463, 0.592755416408181, 0.327831673435867, 0.995317445602268, 0.803048126166686, 0.138109270250425, 0.82894026231952, 0.590680544031784, 0.142904181033373, 0.239317751023918, 0.574796024244279, 0.928915839409456, 0.272088380064815, 0.923303628107533, 0.65078464592807, 0.658562712138519, 0.617850201902911, 0.431175895500928, 0.851965599926189, 0.150527018820867, 0.906041329028085, 0.763294621370733, 0.0741146011278033, 0.0208454760722816, 0.486634212080389, 0.550622413866222, 0.264990088297054, 0.305352095514536, 0.951081259408966, 0.546383751090616, 0.255174739751965, 0.836815159069374, 0.588204660685733, 0.350819974206388, 0.544426277745515, 0.346034562215209, 0.533646404743195, 0.779381994623691, 0.973915243055671, 0.246561407810077, 0.88505594804883, 0.0614174681250006, 0.679351365892217, 0.485066212480888, 0.373153993394226, 0.131963251857087, 0.160339909605682, 0.504627715796232, 0.783948387484998, 0.649842882994562, 0.348023809725419, 0.410793588962406, 0.579820923507214, 0.461240133270621, 0.245820871088654, 0.819567438680679, 0.84648656938225, 0.0744207762181759, 0.731086306273937, 0.391778159653768, 0.386343380901963, 0.225103244185448, 0.659803349757567), actual = c(0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"