Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.2064724        0.2924578 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5578369        0.5578369        0.5029694 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.406530357897282, 0.0705207374412566, 0.905613062437624, 0.711933846818283, 0.368743092054501, 0.0178812702652067, 0.741178360069171, 0.705952757271007, 0.125464358599856, 0.665191532112658, 0.832641773158684, 0.455835211090744, 0.246733186999336, 0.515236348146573, 0.0507177808322012, 0.299544502282515, 0.58706625807099, 0.496468377299607, 0.785656521795318, 0.933663251809776, 0.177862716140226, 0.10919430712238, 0.573079632362351, 0.533408474409953, 0.649398681940511, 0.179913510335609, 0.485236146487296, 0.810958281625062, 0.829886066261679, 0.183212448609993, 0.658980888780206, 0.39040808333084, 0.604866002453491, 0.922802614048123, 0.385532149113715, 0.640949672786519, 0.350102401804179, 0.574750919826329, 0.374340692302212, 0.0745399659499526, 0.938041258603334, 0.480342473834753, 0.180385068291798, 0.508872100617737, 0.211646287236363, 0.488532272167504, 0.39691781019792, 0.792602020781487, 0.166785796172917, 0.0526895667426288, 0.714082609163597, 0.153466489166021, 0.923228471074253, 0.894407750573009, 0.0556839010678232, 0.0549704185687006, 0.876089610392228, 0.420169166987762, 0.0172867788933218, 0.612718833610415, 0.438255702611059, 0.27040407480672, 0.879870972130448, 0.412056368310004, 0.336228199535981, 0.367903211619705, 0.0374253592453897, 0.51795577048324, 0.352413653861731, 0.722955849720165, 0.477457197150216, 0.694914241554216, 0.218417062656954, 0.763730878243223, 0.20247470936738, 0.214336328906938, 0.326802856754512, 0.977987005375326, 0.567830530460924, 0.354365386068821, 0.608943011844531, 0.90454940055497, 0.995615585939959, 8.52285884320736e-05, 0.445208332501352, 0.349960461026058, 0.548017578665167, 0.121463768882677, 0.221865197876468, 0.654424561653286, 0.878710958873853, 0.255868601845577, 0.245714128017426, 0.0413475935347378, 0.416097189532593, 0.792584357317537, 0.311906705843285, 0.188073334982619, 0.973868335364386, 0.0804309574887156, 0.907655903836712, 0.970570337260142, 0.928638509009033, 0.184453525347635, 0.759605168830603, 0.981278487946838, 0.891019668895751, 0.579254631884396, 0.114946816349402, 0.991681825369596, 0.466765143210068, 0.814804775174707, 0.379176945891231, 0.514850117033347, 0.878605401841924, 0.216253825696185, 0.416039071511477, 0.390844194218516, 0.535266599385068, 0.0644765545148402, 0.17548521887511, 0.9685533803422, 0.69596457364969, 0.0573996126186103, 0.999691231641918, 0.480187339242548, 0.387479098513722, 0.784961844561622, 0.855815920745954, 0.461749357404187, 0.061088518705219, 0.144270776305348, 0.917789130238816, 0.197157907532528, 0.409639217890799, 0.668129557976499, 0.397598857991397, 0.0867577553726733, 0.165871429489926, 0.658820364158601, 0.290550481760874, 0.438531832303852, 0.348966187797487, 0.50226275017485, 0.833781336201355, 0.670516092563048, 0.422119292197749, 0.851229408755898, 0.053350361995399, 0.545346457511187, 0.39322421210818, 0.174147037090734, 0.716904334723949, 0.269308123271912, 0.549778134794906, 0.86250276514329, 0.360615058336407, 0.163235268322751, 0.950149012496695, 0.480132534867153, 0.55528535367921, 0.29518268047832, 0.669151103822514, 0.398534797597677, 0.888122159987688, 0.17861410183832, 0.469309739535674, 0.886102410499007, 0.838386549148709, 0.856676845811307, 0.276737307198346, 0.751960817258805, 0.49647129746154, 0.414469928247854, 0.687627848237753, 0.504890743410215, 0.0654497467912734, 0.858568442752585, 0.7164894239977, 0.0870660836808383, 0.746093649882823, 0.639645168092102, 0.154678944032639, 0.306644510477781, 0.770936211105436, 0.234310369007289, 0.23041174467653, 0.641212276415899, 0.142372202593833, 0.192694335477427, 0.901137099368498, 0.239019728265703, 0.468958762707189, 0.265371708199382, 0.988876507617533, 0.362780595198274, 0.950970195000991, 0.782308415975422, 0.366950512165204, 0.754405561834574, 0.0993445902131498, 0.478131658630446, 0.461092984769493, 0.0525956465862691, 0.984504049411044, 0.384090226376429, 0.0740060326643288, 0.857998383929953, 0.399780116276816, 0.0100640126038343, 0.756974168820307, 0.0312179999891669, 0.148002971429378, 0.967713694553822, 0.993504193844274, 0.443386680213735, 0.100850564194843, 0.693103089928627, 0.305401676800102, 0.921455263160169, 0.638624014565721, 0.733852613484487, 0.64318343764171, 0.876161921769381, 0.0742887265514582, 0.0917048158589751, 0.278356074821204, 0.886804015841335, 0.106560417450964, 0.900503810727969, 0.976510700769722, 0.544458300573751, 0.489116365555674, 0.270015018060803, 0.5727707920596, 0.346755811013281, 0.769096295116469, 0.177890758030117, 0.316419901559129, 0.0998046260792762, 0.282192833954468, 0.507681683870032, 0.597502385498956, 0.140048980945721, 0.880100389942527, 0.0629673397634178, 0.766823729965836, 0.600965675897896, 0.498504027025774, 0.0868848639074713, 0.988656521309167, 0.736770977964625, 0.977685654070228, 0.311576512409374, 0.893061565700918, 0.150971277151257, 0.70841219718568, 0.396164575591683, 0.838106608949602, 0.526390144368634, 0.800355219049379, 0.724999880883843, 0.867784664034843, 0.169751249719411, 0.541592641500756, 0.567009993828833, 0.948527722153813, 0.79747953126207, 0.779465828323737, 0.342275221366435, 0.490501638967544, 0.0384862886276096, 0.647880130214617, 0.108850351534784, 0.458519004983827, 0.996229821583256, 0.607406962430105, 0.203442725585774, 0.845114667201415, 0.678358677774668, 0.859051735838875, 0.52731116884388, 0.87973780115135, 0.383608988719061, 0.454450331395492, 0.258037309162319, 0.855186929693446, 0.907428230158985, 0.859118762658909, 0.0981951428111643, 0.304888331796974, 0.352613066788763, 0.585992755834013, 0.174826913513243, 0.968355519929901, 0.181272484129295, 0.232370350975543, 0.721781883854419, 0.792383641935885, 0.915899718413129, 0.847569665405899, 0.483040808700025, 0.972778275609016, 0.868038854328915, 0.66932271560654, 0.968268186319619, 0.937029905151576, 0.348572355229408, 0.349869540194049, 0.502621744060889, 0.191180283436552, 0.492177519248798, 0.3782222061418, 0.284009434748441, 0.167172219138592, 0.177235797513276, 0.409447904210538, 0.750825218390673, 0.656370509881526, 0.614721560152248, 0.808099873596802, 0.960760712856427, 0.0573355106171221, 0.820065308827907, 0.804603402735665, 0.41179385734722, 0.290958081139252, 0.751818471588194, 0.172087196959183, 0.768097530351952, 0.379026303300634, 0.82747091143392, 0.471749164164066, 0.448430554009974, 0.628054237226024, 0.67165285651572, 0.420359723269939, 0.0311373246368021, 0.837275934638456, 0.935733502730727, 0.121848635142669, 0.725517081329599, 0.817749575478956, 0.0884285906795412, 0.772009355714545, 0.98868469055742, 0.60391039121896, 0.59145148168318, 0.322810134850442, 0.161894524469972, 0.412108243210241, 0.580692415125668, 0.393189013469964, 0.272724225884303, 0.581202099798247, 0.326567733893171, 0.442687313538045, 0.722920250147581, 0.109244641615078, 0.187062332639471, 0.229863188927993, 0.958640562137589, 0.932846011826769, 0.904178362339735, 0.72842194722034, 0.481362970545888, 0.972416104748845, 0.63097850442864, 0.233852020697668, 0.451415721559897, 0.346038994379342, 0.983041361439973, 0.540004353970289, 0.434511355357245, 0.289465277688578, 0.737348537659273, 0.298832940403372, 0.527145440690219, 0.810640024254099, 0.355325613170862, 0.403277300531045, 0.656182921258733, 0.97877454967238, 0.636039654491469, 0.473133144667372, 0.366914543323219, 0.43837555940263, 0.382701793918386, 0.0791332046501338, 0.0381310693919659, 0.695368988439441, 0.929054118227214, 0.520997469546273, 0.195692773209885, 0.163051356794313, 0.536834177561104, 0.885647653369233, 0.216308050788939, 0.645981996785849, 0.111501617357135, 0.760185932042077, 0.366533265449107, 0.0241409183945507, 0.724542989395559, 0.438393475487828, 0.330162542173639, 0.531181201338768, 0.662543034879491, 0.226930265547708, 0.642283296445385, 0.604369689710438, 0.842149984324351, 0.325626821490005, 0.618845290271565, 0.805244568735361, 0.316980341915041, 0.928328194655478, 0.0478690611198545, 0.776912773260847, 0.824222379829735, 0.428637109231204, 0.871842102147639, 0.513129592873156, 0.0168301216326654, 0.949277862906456, 0.470846388256177, 0.0147891016677022, 0.410611979430541, 0.136599567020312, 0.717579355696216, 0.0959183019585907, 0.41610343195498, 0.224454239709303, 0.242189258569852, 0.535326782381162, 0.72441074391827, 0.763126281090081, 0.941649231128395, 0.844466031529009, 0.174981266260147, 0.378155727637932, 0.945117542985827, 0.870805972954258, 0.542101653758436, 0.0610314020887017, 0.421871059341356, 0.0556987866293639, 0.843436605995521, 0.462171474006027, 0.474003300769255, 0.570751049555838, 0.598857532721013, 0.41430277749896, 0.417658092454076, 0.364519549999386, 0.410006472840905, 0.296104430221021, 0.115342336241156, 0.31589946616441, 0.57667253119871, 0.904409199487418, 0.688561047660187, 0.943861127365381, 0.548445267369971, 0.56722366833128, 0.0719307458493859, 0.511115243425593, 0.935113871935755, 0.192399821942672, 0.646908384747803, 0.939669742714614, 0.164586388505995, 0.924146447563544, 0.804076276021078, 0.971673431340605, 0.116684435168281, 0.379221954382956, 0.707784378901124, 0.377448746003211, 0.39585690991953, 0.148493953282014, 0.118944415589795, 0.0552417852450162, 0.826715446077287, 0.690356394508854, 0.0873276486527175, 0.758638835512102, 0.449155099224299, 0.630357339512557, 0.457484635990113, 0.559365197550505, 0.26571821118705, 0.674902765080333, 0.996405100217089, 0.894397330004722, 0.797759741544724, 0.893250775989145, 0.703406498767436, 0.832150779664516, 0.52495187567547, 0.706748996162787, 0.689196224790066, 0.0134906766470522, 0.965205734595656, 0.451754416339099, 0.861104835756123, 0.595824313117191, 0.633225591154769, 0.421438542194664, 0.925330952042714, 0.293169638607651, 0.585435556015, 0.934515132801607, 0.467315612360835, 0.631097837584093, 0.836852243868634, 0.106533990707248, 0.0687672710046172, 0.788571833400056, 0.82484318385832, 0.631496706744656, 0.814802437089384, 0.798940513283014, 0.336349444463849, 0.0721336579881608, 0.588382553542033, 0.261595700168982, 0.656837886897847, 0.0336782762315124, 0.67919581476599, 0.0657331480178982, 0.154904911760241, 0.270167955430225, 0.944267804501578, 0.182763936230913, 0.759950399864465, 0.820972286164761, 0.684155934024602, 0.613287881715223, 0.745721789076924, 0.433429111959413, 0.741279761539772, 0.282426831079647, 0.762558755464852, 0.714837568346411, 0.986255209892988, 0.085601391736418, 0.693138906499371, 0.105050839018077, 0.949451594846323, 0.331166384276003, 0.973167447606102, 0.195265810005367, 0.185742074623704, 0.0179874992463738, 0.046621385961771, 0.573199597653002, 0.791138257598504, 0.992275927215815, 0.0464200272690505, 0.39383157948032, 0.718675158685073, 0.882429631659761, 0.348229649243876, 0.550410815514624, 0.368690375937149, 0.286853636149317, 0.926328326808289, 0.234886044636369, 0.345297025283799, 0.499725848203525, 0.267312719486654, 0.875644164392725, 0.323167786700651, 0.856321799801663, 0.895989687880501, 0.124620833666995, 0.237644577864558, 0.148717550328001, 0.589855725876987, 0.136455006897449, 0.966107381740585, 0.74366162950173, 0.276140007190406, 0.631170053500682, 0.101554665481672, 0.573764922097325, 0.515440116869286, 0.0671924941707402, 0.358284641755745, 0.0801977657247335, 0.340487922308967, 0.568993126275018, 0.0091027575545013, 0.0208611490670592, 0.643488629488274, 0.0759835646022111, 0.586615010397509, 0.973850990179926, 0.22090830327943, 0.941867272835225, 0.681158617604524, 0.356399297248572, 0.805515125859529, 0.446953258942813, 0.878330497071147, 0.00227456400170922, 0.378369328565896, 0.741754690650851, 0.209406530950218, 0.363369001075625, 0.403584430459887, 0.877736762631685, 0.566790008218959, 0.935844680294394, 0.219160298584029, 0.26246434240602, 0.36508425604552, 0.29530803184025, 0.253437964478508, 0.978713609511033, 0.920588723383844, 0.457758071599528, 0.7912345782388, 0.324958260171115, 0.401884680846706, 0.526107210433111, 0.488445741124451, 0.686783353099599, 0.775479659903795, 0.850628256332129, 0.575355070643127, 0.593451937660575, 0.320420960430056, 0.450247769476846, 0.745771642774343, 0.936102248262614, 0.64405518816784, 0.912535817828029, 0.251566974213347, 0.762253938242793, 0.858741886215284, 0.580789889674634, 0.834842026466504, 0.788330224109814, 0.126900386530906, 0.125410983571783, 0.649786776630208, 0.334142670501024, 0.034051023889333, 0.974931447533891, 0.993325363844633, 0.0633643618784845, 0.42498330399394, 0.156415965873748, 0.960946430452168, 0.128727624891326, 0.292738527059555, 0.598870879970491, 0.729532529832795, 0.791072049643844, 0.541490063769743, 0.942412163829431, 0.290868705837056, 0.0874406527727842, 0.603181090205908, 0.890801482368261, 0.273318256950006, 0.660076154628769, 0.853557865368202, 0.1756353052333, 0.765585104236379, 0.0873005078174174, 0.983643038198352, 0.365389324491844, 0.860834662569687, 0.663329585222527, 0.236074453918263, 0.0125471316277981, 0.802857550093904, 0.0916366991586983, 0.0501704132184386, 0.671097436919808, 0.404474087059498, 0.0652065947651863, 0.981625792104751, 0.922203497262672, 0.889435598626733, 0.731995566282421, 0.642594656441361, 0.928641718113795, 0.239715109346434, 0.809033464407548, 0.466784595046192, 0.439976963214576, 0.198280860902742, 0.488043377641588, 0.508710659574717, 0.470696929842234, 0.306387801887468, 0.445662614190951, 0.621480894042179, 0.526031861780211, 0.335855711717159, 0.122536561917514, 0.788715076632798, 0.359824985964224, 0.742688235128298, 0.493967207847163, 0.246627776185051, 0.590770477661863, 0.757979395100847, 0.594309086212888, 0.326687507331371, 0.244741340633482, 0.258170170942321, 0.990872008958831, 0.299911226145923, 0.859522312413901, 0.653392655076459, 0.735741003183648, 0.524827021174133, 0.911371380090714, 0.642643311060965, 0.477159743662924, 0.790969991125166, 0.32961358060129, 0.769570637959987, 0.25000092596747, 0.387907552300021, 0.0272494778037071, 0.481942160055041, 0.771474597044289, 0.866635535378009, 0.0475039056036621, 0.33453546767123, 0.912513487040997, 0.141893497901037, 0.848099175374955, 0.108371625421569, 0.898270717123523, 0.993494633352384, 0.731632373295724, 0.587014223681763, 0.59901143447496, 0.52687302348204, 0.359607056947425, 0.109100585337728, 0.741471673129126, 0.869926182087511, 0.463477763347328, 0.675844046985731, 0.727222190704197, 0.303794594947249, 0.325531880836934, 0.439881105907261, 0.285558727802709, 0.983631833922118, 0.880081447772682, 0.690188082866371, 0.575615097070113, 0.370132470969111, 0.445243182824925, 0.299011980183423, 0.0760456072166562, 0.633165488718078, 0.0479786347132176, 0.199625268578529, 0.560814850963652, 0.213769435882568, 0.479058128083125, 0.856457968708128, 0.937080160947517, 0.595964434789494, 0.920202999841422, 0.0708117573522031, 0.934014387661591, 0.269284777343273, 0.317349254852161, 0.666825661668554, 0.118290859507397, 0.820357826305553, 0.0372440840583295, 0.581408711150289, 0.224779046839103, 0.363838983001187, 0.900461491663009, 0.905250840354711, 0.830414762021974, 0.711765639251098, 0.830856990301982, 0.13080556737259, 0.938214367022738, 0.0733164753764868, 0.375166427344084, 0.793574845883995, 0.77666381932795, 0.923892400227487, 0.217584648169577, 0.496804614784196, 0.05225022835657, 0.452711296966299, 0.849397389683872, 0.267918856348842, 0.310821182793006, 0.197043646592647, 0.558396730804816, 0.356820586370304, 0.263081495184451, 0.241258634719998, 0.0296710964757949, 0.879974480951205, 0.444740823237225, 0.274489559000358, 0.85275203967467, 0.907600230537355, 0.363783224718645, 0.16846955451183, 0.802993277553469, 0.782580443192273, 0.994730409234762, 0.958000878803432, 0.0803567366674542, 0.780735280597582, 0.769712596898898, 0.563885802868754, 0.80432785442099, 0.117630788125098, 0.507577535696328, 0.297383261611685, 0.385434898082167, 0.83496742323041, 0.810704348841682, 0.41382146673277, 0.700916168978438, 0.075678980210796, 0.782588096568361, 0.841897988691926, 0.478184390813112, 0.659139941213652, 0.878529048990458, 0.824949085945264, 0.9096921288874, 0.897643693257123, 0.436136201955378, 0.678349884459749, 0.596270041773096, 0.823263114783913, 0.63858682080172, 0.6866519106552, 0.248039421159774, 0.461378342704847, 0.172699273563921, 0.80687544005923, 0.821713672019541, 0.431688542943448, 0.704783407505602, 0.14711165940389, 0.920175789156929, 0.77507783100009, 0.186111117713153, 0.837837515166029, 0.821898907423019, 0.884251966606826, 0.756061480613425, 0.798658423125744, 0.790842669084668, 0.753202112391591, 0.693713550223038, 0.365908444626257, 0.059024894842878, 0.867175317369401, 0.136798795778304, 0.471557356417179, 0.914563625352457, 0.807431635912508, 0.0885679062921554, 0.458429991733283, 0.813879621215165, 0.0534863730426878, 0.697425266960636, 0.472885468276218, 0.272258399520069, 0.055997067829594, 0.63335336977616, 0.649715093430132, 0.844476487720385, 0.338583477772772, 0.84427582169883, 0.978093089535832, 0.0413196857552975, 0.929390216944739, 0.302938035223633, 0.560979950241745, 0.0497311444487423, 0.461758570978418, 0.619287830544636, 0.3542331433855, 0.690546571277082, 0.863241259474307, 0.0111116315238178, 0.150179940043017, 0.695183650590479, 0.52978052594699, 0.323385541327298, 0.297625398496166, 0.515683444682509, 0.621841687476262, 0.516172987874597, 0.304769941838458, 0.571984197711572, 0.588802149752155, 0.958997451234609, 0.139402735978365, 0.829785575391725, 0.916990639874712, 0.223894926952198, 0.549964118050411, 0.839980867691338, 0.682066956534982, 0.436987952096388, 0.872027172474191, 0.728665946982801, 0.0568205898161978, 0.849667041329667, 0.296077814884484, 0.131770985899493, 0.168109802296385, 0.957685756264254, 0.0184068134985864, 0.979798312764615, 0.371807758929208, 0.126597719499841, 0.254927753936499, 0.996995914261788, 0.636374050052837, 0.481106608174741, 0.624335273401812, 0.409019866259769, 0.236418717773631, 0.14207182591781, 0.745683202054352, 0.0858743828721344, 0.222951284609735, 0.651342512341216, 0.669069015420973, 0.313132552895695, 0.409934650408104, 0.966760700335726, 0.825969280907884, 0.543363362317905, 0.316735942382365, 0.764518548734486, 0.227820308180526, 0.654666398884729, 0.656157416757196, 0.483536010608077, 0.552405228838325, 0.834569316590205, 0.870153984054923, 0.498082651989534, 0.142987502040341, 0.984469121322036, 0.606517560314387, 0.228386329254135, 0.188256488647312, 0.49982429575175, 0.292373939650133, 0.672713159117848, 0.356672884197906, 0.264711201423779, 0.0141347912140191, 0.519459653645754, 0.205479891737923, 0.433218437014148, 0.799027094617486, 0.715270636836067, 0.321827216073871, 0.420312832808122, 0.909594703000039, 0.895763223292306, 0.167533698724583, 0.479793067788705, 0.305852328427136, 0.317577425390482, 0.0191413897555321, 0.37833901681006, 0.127587543567643, 0.428548757918179, 0.15069695096463, 0.899783137254417, 0.80625585000962, 0.775442197453231, 0.240955455694348, 0.551779108820483, 0.671633162302896, 0.320160661591217, 0.743573885876685, 0.149952163919806, 0.345029686577618), actual = c(1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.288672046037391, 0.609547753818333, 0.474952878430486, 0.187118313042447, 0.126475669909269, 0.75246148230508, 0.499469841597602, 0.455146247288212, 0.771954091265798, 0.201398276956752, 0.374409331707284, 0.164336453191936, 0.0441105321515352, 0.949844048591331, 0.618515112204477, 0.777211987879127, 0.910197786288336, 0.410564164863899, 0.178604040760547, 0.499989313073456, 0.365648301085457, 0.988115838728845, 0.906206078128889, 0.0955770085565746, 0.110077050980181, 0.218104850267991, 0.880363957723603, 0.112536678789183, 0.202353907283396, 0.821595364250243, 0.308591921348125, 0.465287459781393, 0.119804877787828, 0.969568240689114, 0.158018844434991, 0.974254766479135, 0.254548714263365, 0.517715095542371, 0.999733319506049, 0.788747578393668, 0.16762686567381, 0.57835912075825, 0.412304722471163, 0.539400813635439, 0.293147507822141, 0.781296802684665, 0.373010490089655, 0.0608345789369196, 0.907592191593722, 0.147103546420112, 0.257206964539364, 0.779142022365704, 0.556902572279796, 0.588842008728534, 0.397968584205955, 0.291925339959562, 0.15671441424638, 0.627586330287158, 0.204460729379207, 0.420145102543756, 0.50018198764883, 0.136918375268579, 0.644535623956472, 0.534984794445336, 0.436617800267413, 0.37174992938526, 0.772925755009055, 0.520813252078369, 0.745908222394064, 0.0470569406170398, 0.132114813895896, 0.486394764622673, 0.551071137888357, 0.432949155103415, 0.0129863808397204, 0.963541375705972, 0.300930102122948, 0.151468449272215, 0.228084095055237, 0.803082134807482, 0.859551468165591, 0.802220262121409, 0.551314946496859, 0.787869861815125, 0.538434657035396, 0.845421381760389, 0.0137919257394969, 0.045536955120042, 0.233119679614902, 0.999634859384969, 0.643241694197059, 0.890157289337367, 0.380512515082955, 0.890075615374371, 0.865738662658259, 0.520949053578079, 0.160357888555154, 0.0431747827678919, 0.850198338041082, 0.151003933744505, 0.869240431813523, 0.0721151120960712, 0.955410819966346, 0.658618696033955, 0.92631059139967, 0.965562989935279, 0.557431874796748, 0.679648638004437, 0.73905253270641, 0.623150302562863, 0.446770965354517, 0.567956340499222, 0.970400708494708, 0.63587318896316, 0.828369892667979, 0.85649056523107, 0.923012226587161, 0.482352452585474, 0.933351002866402, 0.77579007204622, 0.932987102307379, 0.680931849405169, 0.957918068859726, 0.806490811752155, 0.428104761056602, 0.124778220895678, 0.805866922950372, 0.725309177068993, 0.753654958447441, 0.941412706160918, 0.348100391915068, 0.870809719664976, 0.606590232113376, 0.411183029180393, 0.440160979982466, 0.449875072110444, 0.403746761381626, 0.685820516664535, 0.959374583559111, 0.06745894276537, 0.696457368787378, 0.00682796048931777, 0.741918532410637, 0.821395672624931, 0.62964259297587, 0.166744920192286, 0.861332118278369, 0.996698089176789, 0.278049263870344, 0.351425173925236, 0.751491389470175, 0.496187685523182, 0.10446633794345, 0.936591831268743, 0.532993570668623, 0.515641635749489, 0.412923817755654, 0.656940159620717, 0.027182292426005, 0.361828627996147, 0.524938252056018, 0.18063179217279, 0.338268315419555, 0.433762618573382, 0.0596921439282596, 0.494740211637691, 0.496659602969885, 0.0142225592862815, 0.338251363951713, 0.141618302557617, 0.311110266018659, 0.126933998661116, 0.919022399000823, 0.0834187832660973, 0.647695004474372, 0.834482194622979, 0.0298767844215035, 0.553070642752573, 0.0174442576244473, 0.76873381040059, 0.234009012347087, 0.620495338691399, 0.8431169425603, 0.16051481012255, 0.153473445447162, 0.0753705005627126, 0.183903183788061, 0.405653796857223, 0.51218462199904, 0.378483967855573, 0.509840453742072, 0.626249783672392, 0.939846991794184, 0.427234638715163, 0.880217906087637, 0.796777887968346, 0.695313927251846, 0.838497311808169, 3.15543729811907e-05, 0.952098784502596, 0.390688009327278, 0.741076951147988, 0.798623085021973, 0.0277238497510552, 0.224898373242468, 0.571031432598829, 0.291749530937523, 0.730815367307514, 0.917020561406389, 0.3445414674934, 0.822911723749712, 0.336089116986841, 0.867848057532683, 0.632226079702377, 0.0308307916857302, 0.304188115987927, 0.312189847230911, 0.653146213619038, 0.0444582160562277, 0.967520905192941, 0.452981127193198, 0.363811010029167, 0.0205564075149596, 0.420811640098691, 0.706399908987805, 0.177377012325451, 0.324843335198238, 0.545016451971605, 0.41843714681454, 0.370285712415352, 0.488030249951407, 0.679128329269588, 0.287049143109471, 0.491595251718536, 0.225990645121783, 0.523448170628399, 0.811609318479896, 0.950517721474171, 0.457353364443406, 0.397452017758042, 0.412887796992436, 0.558794486103579, 0.764798204181716, 0.200993360951543, 0.272283371537924, 0.383748397929594, 0.13450360763818, 0.672254497883841, 0.241366229718551, 0.836807440500706, 0.606480393093079, 0.279908134834841, 0.978109748568386, 0.364219812676311, 0.235903320368379, 0.715796186588705, 0.195904097286984, 0.764516555238515, 0.6883492723573, 0.167570693185553, 0.329651727108285, 0.768617710098624, 0.766119995620102, 0.797769654542208, 0.853897526860237, 0.627786578377709, 0.675642019836232, 0.481791733531281, 0.0260848961770535, 0.64377604983747, 0.333907028660178, 0.0205819967668504, 0.918816373217851, 0.787379392655566, 0.0654831775464118, 0.434315847000107, 0.117181428009644, 0.499925407348201, 0.750850585289299, 0.148834793595597, 0.688949241302907, 0.825928554637358, 0.0588358070235699, 0.990555334836245, 0.552204198902473, 0.890594588359818, 0.551306349691004, 0.441045669373125, 0.455847115488723, 0.397644844371825, 0.26298137428239, 0.842796976910904, 0.570647589629516, 0.290057932958007, 0.513673142297193, 0.709277375601232, 0.825049498351291, 0.966393792303279, 0.851370227523148, 0.524075454100966, 0.138733085244894, 0.173242300748825, 0.960900660837069, 0.612398461904377, 0.999352609273046, 0.806246590800583, 0.744941843906417, 0.644535777159035, 0.311136743985116, 0.80390004790388, 0.771329570095986, 0.554878968978301, 0.584646547678858, 0.0747298269998282, 0.957939594518393, 0.254857301246375, 0.958538462407887, 0.392714133951813, 0.674656199524179, 0.145822487073019, 0.788728483021259, 0.994740335503593, 0.775354372570291, 0.0920652346685529, 0.400441920850426, 0.33379067806527, 0.352029806701466, 0.165826958836988, 0.368542744312435, 0.644718988798559, 0.331060033524409, 0.386863311519846, 0.543812938500196, 0.202898306306452, 0.701192709850147, 0.757867257110775, 0.460609716596082, 0.100593026960269, 0.381090880837291, 0.0368639517109841, 0.859312189975753, 0.666799552738667, 0.92315890872851, 0.705315420404077, 0.00668121525086462, 0.719711378449574, 0.852246813243255, 0.858767934609205, 0.207458524266258, 0.69823517720215, 0.642903217347339, 0.719357000198215, 0.0532143986783922, 0.33936205226928, 0.483716578921303, 0.22923124092631, 0.62146716681309, 0.869758953573182, 0.332496209535748, 0.44956370908767, 0.521264655981213, 0.399499436141923, 0.486521012382582, 0.091734457295388, 0.609585318248719, 0.152938708895817, 0.657119499985129, 0.232513352297246, 0.567627718672156, 0.328578830230981, 0.980049168225378, 0.680924214888364, 0.720526431687176, 0.595521511510015, 0.153269356815144, 0.54178066062741, 0.980476032709703, 0.353898852830753, 0.705758490832523, 0.901660317089409, 0.71805799077265, 0.0923644346185029, 0.671934822807088, 0.754844185197726, 0.774804407265037, 0.232455944176763, 0.991428897948936, 0.828045776346698, 0.73213171842508, 0.819494206691161, 0.655011035501957, 0.0763081631157547, 0.289828915148973, 0.618068045703694, 0.323676663916558, 0.274310763459653, 0.342679721536115, 0.811449151253328, 0.183191258925945, 0.261748980963603, 0.426659226883203, 0.204078986542299, 0.528164679650217, 0.883318996522576, 0.44122930127196, 0.180056691169739, 0.0343301668763161, 0.28851567953825, 0.849092285847291, 0.231452729785815, 0.0730299893766642, 0.672156166750938, 0.420295997522771, 0.0745622864924371, 0.44187508802861, 0.92647083173506, 0.19736102106981, 0.773923161206767, 0.791327924234793, 0.786596766673028, 0.605682250112295, 0.264810691587627, 0.714677644893527, 0.845108228269964, 0.71792146936059, 0.0475449494551867, 0.305980534991249, 0.497418301180005, 0.90641048992984, 0.91264866781421, 0.109003040939569, 0.325152842095122, 0.214046239620075, 0.779539240989834, 0.837727589532733, 0.814399064984173, 0.543526437366381, 0.728971989359707, 0.960316340904683, 0.55091923289001, 0.527630096301436, 0.694375879596919, 0.390800254652277, 0.265059751458466, 0.951622199034318, 0.167037479812279, 0.422667789272964, 0.289087879704311, 0.150785389821976, 0.796690814429894, 0.300794619368389, 0.769036069046706, 0.933546081185341, 0.152971859322861, 0.211326012155041, 0.562842920655385, 0.554832047084346, 0.775248078163713, 0.567367187468335, 0.236608488950878, 0.262104963418096, 0.332255476620048, 0.770631110528484, 0.0167142227292061, 0.454017032170668, 0.333006314234808, 0.0651120552793145, 0.81527853035368, 0.797829939750955, 0.0121358844917268, 0.278283037710935, 0.105375824263319, 0.0699364382307976, 0.456141714472324, 0.947107729734853, 0.217676651431248, 0.283716331701726, 0.0966919446364045, 0.273099252488464, 0.975114512024447, 0.424169785343111, 0.657107093837112, 0.393248067237437, 0.683546993182972, 0.983456783927977, 0.547223079251125, 0.0763642280362546, 0.841479416238144, 0.35134543874301, 0.654053007485345, 0.814385121921077, 0.813556752633303, 0.550441521219909, 0.809890604810789, 0.217041681520641, 0.621737067587674, 0.170809802832082, 0.319153576390818, 0.817318038549274, 0.170918711926788, 0.604741213377565, 0.681507378816605, 0.120139034697786, 0.112487118225545, 0.519961450947449, 0.735976809402928, 0.824920061044395, 0.625938144978136, 0.947143178200349, 0.924987057689577, 0.88421846437268, 0.0243250962812454, 0.38277224986814, 0.801590465242043, 0.984139746753499, 0.733213265892118, 0.435486654052511, 0.283281397772953, 0.794798827264458, 0.521982279373333, 0.757109500467777, 0.971239754697308, 0.236374355619773, 0.681198694044724, 0.879369674017653, 0.0757511775009334, 0.0794015326537192, 0.661238461267203, 0.877292919438332, 0.319484220584854, 0.50441670184955, 0.140703666489571, 0.212596192024648, 0.0692600677721202, 0.242066992679611, 0.835325523978099, 0.19761555059813, 0.856870500138029, 0.804907120997086, 0.0417972141876817, 0.658646983327344, 0.23942109849304, 0.303627924062312, 0.0314135672524571, 0.10786398104392, 0.710581994149834, 0.236327416496351, 0.117434179876, 0.474200350232422, 0.486927886027843, 0.822653343435377, 0.0559240754228085, 0.239883778383955, 0.450608051149175, 0.853763320483267, 0.222390794893727, 0.575278025818989, 0.169249043567106, 0.126243285601959, 0.728210597997531, 0.214031089330092, 0.315510825254023, 0.231595720630139, 0.0109845623373985, 0.360241482732818, 0.686283425660804, 0.00517175905406475, 0.0999656901694834, 0.238161555491388, 0.324029410490766, 0.0163808106444776, 0.993224684847519, 0.439711135579273, 0.424327796557918, 0.270050070481375, 0.174847099231556, 0.845561889233068, 0.127253769896924, 0.257638669805601, 0.789313810411841, 0.299134061904624, 0.248516383115202, 0.624789950205013, 0.0979113278444856, 0.754342312924564, 0.603293729946017, 0.651106429751962, 0.76419164147228, 0.214824137510732, 0.284348451765254, 0.321635590866208, 0.538783329771832, 0.424523031339049, 0.852442925097421, 0.00788659113459289, 0.755818698788062, 0.474145228974521, 0.817049578530714, 0.489926365204155, 0.39579025353305, 0.693075362592936, 0.173110155388713, 0.14047825243324, 0.251275587594137, 0.556968048214912, 0.546988545916975, 0.00749610061757267, 0.329930779756978, 0.715267477557063, 0.225739559391513, 0.248299200320616, 0.778835089178756, 0.528506046161056, 0.694145670160651, 0.949129382614046, 0.381312502548099, 0.862294339342043, 0.155484463786706, 0.263956732815132, 0.744530862895772, 0.52396513428539, 0.318010473623872, 0.964293305529281, 0.362635251134634, 0.200743776280433, 0.732750771567225, 0.661053832154721, 0.413214454893023, 0.68588223378174, 0.583672969136387, 0.46887704427354, 0.544794668676332, 0.715276271570474, 0.855927941156551, 0.331405279692262, 0.48861985700205, 0.168623917968944, 0.729667075909674, 0.111692651640624, 0.0492683320771903, 0.83733719214797, 0.735723028425127, 0.726852292427793, 0.417535175103694, 0.936933071119711, 0.969159174943343, 0.894535318715498, 0.50415526330471, 0.220643202774227, 0.638199872337282, 0.836624626535922, 0.601994769414887, 0.187954401131719, 0.254479118622839, 0.730194560717791, 0.115249785594642, 0.118135428754613, 0.802542470162734, 0.0168564980849624, 0.493705367669463, 0.0449910664465278, 0.0388150080107152, 0.246383242076263, 0.364323892863467, 0.0543909594416618, 0.203755457187071, 0.643867073114961, 0.816868915455416, 0.469962232047692, 0.930931058013812, 0.449065722990781, 0.515814833110198, 0.569568984676152, 0.629703423241153, 0.641170971561223, 0.529527129139751, 0.848327281186357, 0.69141777115874, 0.485783634474501, 0.839959634235129, 0.0435309887398034, 0.122749004513025, 0.833179829176515, 0.236875440226868, 0.33602232253179, 0.616802363656461, 0.112431960180402, 0.208122099516913, 0.651759509230033, 0.0132817786652595, 0.00193649413995445, 0.229759356705472, 0.878376118373126, 0.0702517202589661, 0.100177963729948, 0.984301693737507, 0.93413644656539, 0.369790378026664, 0.415382395731285, 0.624395442428067, 0.420139693422243, 0.618648509029299, 0.575476514641196, 0.618729445151985, 0.311854765983298, 0.49782465887256, 0.43608414218761, 0.504482971038669, 0.514393319841474, 0.578750376123935, 0.717778657563031, 0.58339353906922, 0.840549144195393, 0.592326707439497, 0.831954507622868, 0.291817020624876, 0.825146902352571, 0.458354125265032, 0.905112128471956, 0.528626081766561, 0.719570933142677, 0.5057213595137, 0.520158666884527, 0.942126049892977, 0.297335233772174, 0.383484842255712, 0.747283668722957, 0.370533771812916, 0.450173890916631, 0.323120203800499, 0.388337699929252, 0.741326642921194, 0.708978596841916, 0.95276084728539, 0.471659488976002, 0.421174445888028, 0.937668211990967, 0.95364916161634, 0.186856114305556, 0.234328748658299, 0.802734104916453, 0.307246886193752, 0.885484502883628, 0.486029573716223, 0.101981776533648, 0.864986335393041, 0.336271231528372, 0.702397487126291, 0.90317733050324, 0.937948260223493, 0.293551016831771, 0.612404860323295, 0.771354756085202, 0.126170783303678, 0.0808697319589555, 0.0987638672813773, 0.393085198942572, 0.899762979708612, 0.660811141366139, 0.899263862753287, 0.687721757218242, 0.0622144211083651, 0.942441288847476, 0.58679363084957, 0.209115961566567, 0.905128659447655, 0.884329279651865, 0.686183434678242, 0.100175073137507, 0.259886004962027, 0.123390372376889, 0.875478104688227, 0.484508727444336, 0.624568838858977, 0.415115792769939, 0.983841472538188, 0.368839718867093, 0.0679530366323888, 0.75423658429645, 0.875889231916517, 0.0222083069384098, 0.243503785226494, 0.222933486802503, 0.628021413227543, 0.000648280140012503, 0.728034692816436, 0.93208877579309, 0.27388962963596, 0.87366084381938, 0.435219895793125, 0.278594764415175, 0.541845411760733, 0.589366164989769, 0.747150082373992, 0.807977818418294, 0.10901104635559, 0.872230387292802, 0.935261500766501, 0.0136280287988484, 0.647992430487648, 0.228768256958574, 0.00812107999809086, 0.39282057271339, 0.500092193949968, 0.620799232507125, 0.659359346842393, 0.746757966466248, 0.434628293383867, 0.525589522439986, 0.948621685151011, 0.810472829267383, 0.0908319717273116, 0.486736309481785, 0.207810162799433, 0.828420363599434, 0.179414576850832, 0.943688567960635, 0.0817725630477071, 0.953618190251291, 0.17364590568468, 0.466260899556801, 0.57829822297208, 0.507949190447107, 0.487233719788492, 0.880706006661057, 0.85303070791997, 0.881300404900685, 0.915683159604669, 0.0335852557327598, 0.189386722398922, 0.818563815206289, 0.345678966259584, 0.848988557234406, 0.43317946465686, 0.846959206974134, 0.767210439546034, 0.464454092085361, 0.0432538464665413, 0.764172723982483, 0.0639578627888113, 0.197125301230699, 0.401480379514396, 0.66166540235281, 0.586088322568685, 0.656530083622783, 0.374835940776393, 0.725066291401163, 0.861707830103114, 0.060436692321673, 0.75849514757283, 0.584404382156208, 0.102394533809274, 0.371432211482897, 0.471944539807737, 0.784203053684905, 0.279898879118264, 0.829984415322542, 0.495615621097386, 0.143329601269215, 0.128034390509129, 0.0491324462927878, 0.000430401647463441, 0.784593344200402, 0.819576353533193, 0.981535211903974, 0.735825531184673, 0.580266927368939, 0.160662884125486, 0.534217350883409, 0.730324884643778, 0.359233506256714, 0.924421312753111, 0.467309304978698, 0.165990304900333, 0.528146723983809, 0.978980932850391, 0.336385626578704, 0.331691256957129, 0.0565079981461167, 0.0302864331752062, 0.201579497428611, 0.911587298382074, 0.867635199101642, 0.251194780226797, 0.69923804490827, 0.233370242873207, 0.73921018303372, 0.664773257216439, 0.458447217708454, 0.295208568684757, 0.355656273663044, 0.101503255078569, 0.638736251508817, 0.654575594700873, 0.571800816338509, 0.741003934293985, 0.469581589335576, 0.734123796224594, 0.744618338532746, 0.393919202266261, 0.391921280650422, 0.552829203894362, 0.0435662907548249, 0.765331234084442, 0.118900690460578, 0.962183777242899, 0.573858251329511, 0.957238385919482, 0.0901665729470551, 0.622029279358685, 0.208392708562315, 0.103416283614933, 0.492444083327428, 0.680261193076149, 0.109483285807073, 0.0768913852516562, 0.49336093920283, 0.241557449102402, 0.486910382751375, 0.21811000443995, 0.866426188964397, 0.00139094982296228, 0.709624438779429, 0.222648877417669, 0.542286744574085, 0.39606687030755, 0.548558779759333, 0.845719432458282, 0.971796041820198, 0.211295563960448, 0.993650231743231, 0.156714134849608, 0.381686360109597, 0.215935880551115, 0.712858526268974, 0.359051810344681, 0.661523318383843, 0.993717293255031, 0.322053624549881, 0.412168303737417, 0.103348455624655, 0.390786808915436, 0.205704031977803, 0.902463760925457, 0.957989339018241, 0.0816128142178059, 0.500358146615326, 0.229126111371443, 0.289446811657399, 0.281807305989787, 0.000781619688495994, 0.961720327846706, 0.448718758998439, 0.250364441424608, 0.0133466585539281, 0.821721193380654, 0.871134206419811, 0.942202169680968, 0.841420503566042, 0.903466620715335, 0.358030666131526, 0.208981062052771, 0.655637366231531, 0.233155855210498, 0.382216975325719, 0.594715574756265, 0.335277918260545, 0.164152905577794, 0.0620837430469692, 0.879175858804956, 0.225418789312243, 0.614467965438962, 0.422708865022287, 0.989364997018129, 0.455186122795567, 0.269085160689428, 0.0930443301331252, 0.645263262791559, 0.535458190832287, 0.741117013851181, 0.706820422085002, 0.152653867378831, 0.833124765893444, 0.101029675221071, 0.73592345090583, 0.7678864877671, 0.000940507976338267, 0.435743113746867, 0.526459725107998, 0.702286309329793, 0.130007655592635, 0.0451116650365293, 0.404344068840146, 0.988247122149915, 0.993464030092582, 0.309336706763133, 0.782065304694697, 0.654451178852469, 0.319079885957763, 0.558498656377196), actual = c(1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"