Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.1851803        0.1851803 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4414829        0.4414829        0.5049541 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.4551296827849, 0.225509590702131, 0.674873910844326, 0.168318151729181, 0.63586011249572, 0.365451858611777, 0.37272859713994, 0.411807557567954, 0.203468944877386, 0.0350247663445771, 0.25930324033834, 0.22083124704659, 0.899472419172525, 0.532607338624075, 0.769054736476392, 0.240451616002247, 0.299961272394285, 0.963408141862601, 0.342520076548681, 0.219874318223447, 0.75895892851986, 0.373351598391309, 0.281613420229405, 0.654564831173047, 0.575323446886614, 0.758699630387127, 0.593765755649656, 0.299703398486599, 0.4532595933415, 0.792872332269326, 0.896335321012884, 0.491632464341819, 0.334719075122848, 0.113520441809669, 0.411048982525244, 0.2367773137521, 0.506821489194408, 0.782396139809862, 0.0596808928530663, 0.533805897459388, 0.159624799387529, 0.187922412995249, 0.489915090380237, 0.547859685262665, 0.547357226256281, 0.555232220795006, 0.0959075288847089, 0.694336613174528, 0.849106163717806, 0.635539987124503, 0.771931267809123, 0.778953357134014, 0.125533614540473, 0.612967327004299, 0.840501544298604, 0.972615111386403, 0.230555773945525, 0.761653941590339, 0.185614277143031, 0.802204721374437, 0.900989078916609, 0.551950660301372, 0.455206386512145, 0.676068050554022, 0.211449153954163, 0.88446502876468, 0.132646407233551, 0.671397028490901, 0.951825123978779, 0.37160097504966, 0.473276249598712, 0.444196727825329, 0.596667751902714, 0.723618843359873, 0.395446634385735, 0.770897127920762, 0.318731599254534, 0.284359396900982, 0.972641361877322, 0.371863655745983, 0.663760148454458, 0.21132532809861, 0.924288291949779, 0.667589055607095, 0.404857198009267, 0.977141801966354, 0.646387747954577, 0.853038667701185, 0.188921249238774, 0.668456511572003, 0.522869160864502, 0.0976954537909478, 0.90904112579301, 0.0713415844365954, 0.535592836327851, 0.827107779216021, 0.622861597454175, 0.786737288814038, 0.770404250361025, 0.491946520050988, 0.104591342387721, 0.824001909000799, 0.296401513274759, 0.647999409586191, 0.161732560954988, 0.607454852666706, 0.80685907532461, 0.890408923150972, 0.663613175740466, 0.856336348690093, 0.858435925329104, 0.886621578363702, 0.312235403805971, 0.318781155860052, 0.0423585341777653, 0.283655426697806, 0.398397034034133, 0.360311519587412, 0.248784371418878, 0.912096009822562, 0.227190740872175, 0.935080813243985, 0.481660491321236, 0.875212612329051, 0.838916328502819, 0.199044654378667, 0.015699588926509, 0.0628274204209447, 0.445912025636062, 0.33808086742647, 0.415395239600912, 0.902069645700976, 0.428501600865275, 0.815707714296877, 0.127575604245067, 0.212592152645811, 0.736640886869282, 0.255024600541219, 0.497380030574277, 0.295285367174074, 0.54432700551115, 0.0398324402049184, 0.0359970482531935, 0.613956360379234, 0.0189720278140157, 0.882340475916862, 0.201750436099246, 0.270271137356758, 0.142872341442853, 0.900783984689042, 0.773502752650529, 0.889428172493353, 0.867734713014215, 0.371543534332886, 0.0584297711029649, 0.736136689083651, 0.906929227756336, 0.341813780134544, 0.026884809602052, 0.525536051020026, 0.0558515409938991, 0.426890013739467, 0.683854529866949, 0.455356250982732, 0.748444995144382, 0.394334436394274, 0.700387540040538, 0.265205587260425, 0.626760051120073, 0.330962277250364, 0.509457802399993, 0.319054550956935, 0.112859508721158, 0.269503833260387, 0.195736695313826, 0.191226158523932, 0.0758377057500184, 0.600145221222192, 0.210919642820954, 0.064872802933678, 0.148374707438052, 0.0107774287462234, 0.71300397766754, 0.24174603424035, 0.051868176786229, 0.772732072975487, 0.420690341154113, 0.991836616536602, 0.0876514336559922, 0.061191163957119, 0.559303789632395, 0.0409875733312219, 0.345395012293011, 0.715598086360842, 0.973011380759999, 0.544212876819074, 0.825237551238388, 0.972931233001873, 0.73709104093723, 0.770079956157133, 0.715496342396364, 0.995558561058715, 0.719818413024768, 0.596408895449713, 0.437566170236096, 0.2024411407765, 0.31147305900231, 0.624836411094293, 0.496938819764182, 0.806137908715755, 0.445094753988087, 0.914171431679279, 0.209946423768997, 0.532077762763947, 0.077278787503019, 0.336985585512593, 0.0862869028933346, 0.54713134188205, 0.756205084966496, 0.556914236862212, 0.696995010133833, 0.0923386546783149, 0.188059881562367, 0.360489020124078, 0.798714684555307, 0.729686338454485, 0.344935335218906, 0.858443186385557, 0.473772915545851, 0.855968440650031, 0.986755331978202, 0.335401487071067, 0.0559643725864589, 0.555998829426244, 0.630805091233924, 0.552583191310987, 0.400309180608019, 0.435362665448338, 0.307132313493639, 0.320778299821541, 0.927799200406298, 0.781106594251469, 0.345943844411522, 0.0997765935026109, 0.45550717250444, 0.0223915837705135, 0.563030268531293, 0.959147564135492, 0.339042663341388, 0.954901712248102, 0.529536893125623, 0.784316909266636, 0.323736675316468, 0.591798237990588, 0.476273122476414, 0.274818351492286, 0.695690419524908, 0.852323852712288, 0.637312446953729, 0.50509889353998, 0.702357834670693, 0.134178463835269, 0.838779514888301, 0.64624790661037, 0.772711115190759, 0.137580675538629, 0.383506850805134, 0.96347726485692, 0.666755644604564, 0.757074994733557, 0.699186366749927, 0.368791569722816, 0.662631008774042, 0.405262097949162, 0.0694371599238366, 0.776420998387039, 0.404469535453245, 0.316144406795502, 0.0839029347989708, 0.0149403435643762, 0.492160274647176, 0.930959126912057, 0.302541202167049, 0.850908352527767, 0.387920814566314, 0.406663419678807, 0.265681242570281, 0.249335535569116, 0.205811185296625, 0.517230694182217, 0.33295894251205, 0.940249091945589, 0.520854927366599, 0.74823269364424, 0.0241887615993619, 0.274058442329988, 0.0333106946200132, 0.0633744886144996, 0.208072794368491, 0.0477863510604948, 0.14211787446402, 0.452380975242704, 0.475418908521533, 0.410513136768714, 0.0224952362477779, 0.4511011056602, 0.323125775437802, 0.197982345474884, 0.827892052708194, 0.63973474688828, 0.678579739062116, 0.818478080676869, 0.417496253503487, 0.2539469122421, 0.936434875940904, 0.682449001818895, 0.155086771352217, 0.231961227254942, 0.0701352958567441, 0.478334364714101, 0.149124423507601, 0.129130451008677, 0.796745272586122, 0.833448145072907, 0.91130000189878, 0.509852362098172, 0.271552548743784, 0.411906912922859, 0.473186165327206, 0.32890940643847, 0.20519686373882, 0.0799415672663599, 0.460224639391527, 0.190633554244414, 0.529521877178922, 0.765458542853594, 0.7173509048298, 0.76392706669867, 0.358492217026651, 0.0525880500208586, 0.109957437962294, 0.878796610515565, 0.262319255620241, 0.892597425729036, 0.645154316909611, 0.231793560320511, 0.768540258286521, 0.346651767380536, 0.637557402253151, 0.0657319203019142, 0.111010430147871, 0.332079509273171, 0.0618773102760315, 0.777348245261237, 0.590133252320811, 0.461232971865684, 0.12863917206414, 0.938503526849672, 0.318366683786735, 0.568787380121648, 0.520647032186389, 0.133879623841494, 0.409522496629506, 0.573157918872312, 0.288840992376208, 0.598264306550846, 0.833860555430874, 0.0469419723376632, 0.0479862657375634, 0.553649800829589, 0.848993642721325, 0.832100863102823, 0.852244885638356, 0.255423298105597, 0.873561271931976, 0.392264892579988, 0.183703530114144, 0.557831169804558, 0.195889701368287, 0.892680953256786, 0.496123002842069, 0.20103116473183, 0.126637214794755, 0.0927113995421678, 0.986459653126076, 0.445314584765583, 0.853445130167529, 0.87362585333176, 0.537221226608381, 0.765616699587554, 0.773232222534716, 0.317837398964912, 0.478658044710755, 0.0926194449421018, 0.883373924065381, 0.620900474954396, 0.0748742162249982, 0.956715987063944, 0.303644792875275, 0.66692634858191, 0.811780549585819, 0.147627130150795, 0.089538378175348, 0.805945822270587, 0.422298385994509, 0.17772647109814, 0.999620386399329, 0.397650693077594, 0.461697355145589, 0.0837456632871181, 0.702085325261578, 0.811650663148612, 0.899529299931601, 0.935909240040928, 0.28100392408669, 0.752809021389112, 0.178811934310943, 0.0339801625814289, 0.188951549120247, 0.711350599536672, 0.913661000085995, 0.0366913394536823, 0.167474050540477, 0.372362048830837, 0.0462244926020503, 0.0621387655846775, 0.0453886680770665, 0.299148806836456, 0.294073393801227, 0.600630482193083, 0.2201904701069, 0.836661356966943, 0.62114384979941, 0.560616228030995, 0.890215712599456, 0.323600525734946, 0.444367871619761, 0.471307914704084, 0.0194940946530551, 0.248142199125141, 0.776884114835411, 0.787033122964203, 0.506677440367639, 0.787152642384171, 0.3250510643702, 0.842199248028919, 0.913712741807103, 0.154529606224969, 0.996131643187255, 0.939804795198143, 0.938790147192776, 0.00418442604131997, 0.771838383749127, 0.02787772170268, 0.24901657178998, 0.181958281667903, 0.437662286916748, 0.336522877682, 0.300785250496119, 0.0470073125325143, 0.269799010362476, 0.506357767619193, 0.763111737556756, 0.937331828288734, 0.661942199571058, 0.812614657217637, 0.162779106991366, 0.849234038265422, 0.210656181210652, 0.836076475214213, 0.749653237406164, 0.0807127940934151, 0.480748105561361, 0.339018313912675, 0.30762556148693, 0.498301036423072, 0.124141175765544, 0.468381176237017, 0.930595488054678, 0.37264905218035, 0.240952186752111, 0.78759412583895, 0.135433366289362, 0.219272057525814, 0.134898198768497, 0.79003321682103, 0.141587109537795, 0.592527271714061, 0.686871737474576, 0.231047281762585, 0.47212826507166, 0.157709192717448, 0.741784468060359, 0.457309520803392, 0.915156144881621, 0.318952817469835, 0.67264164192602, 0.633714042371139, 0.789331347215921, 0.794399953447282, 0.38884669309482, 0.749579874332994, 0.150359181454405, 0.982991597848013, 0.835097191855311, 0.649677158566192, 0.835287490859628, 0.854984781239182, 0.601134340045974, 0.791814758908004, 0.219370570266619, 0.358944647479802, 0.907648541033268, 0.473839764250442, 0.714463563635945, 0.656809481792152, 0.261348457075655, 0.297107336577028, 0.414978410350159, 0.322132648201659, 0.912430571159348, 0.519377036951482, 0.0984313464723527, 0.255983561044559, 0.0708325891755521, 0.827757314778864, 0.136625293642282, 0.715748079121113, 0.652593591250479, 0.380464497255161, 0.803768231999129, 0.244656194699928, 0.323634434724227, 0.605628102086484, 0.210415376350284, 0.0636572139337659, 0.107157836668193, 0.127241808921099, 0.47337267594412, 0.13158158888109, 0.68252234486863, 0.591181470779702, 0.397617685375735, 0.608523488044739, 0.103874826570973, 0.185141107067466, 0.345171588473022, 0.331547468202189, 0.498179460410029, 0.690245226258412, 0.558517438126728, 0.884570825612172, 0.443301997380331, 0.346008243970573, 0.731098919641227, 0.431004282552749, 0.790420798584819, 0.995403428561985, 0.247618298279122, 0.509706028737128, 0.459321457892656, 0.0412467357236892, 0.224788109771907, 0.0965942160692066, 0.689241415122524, 0.586842676857486, 0.740196395199746, 0.601104117231444, 0.548115953337401, 0.797215410042554, 0.502316620433703, 0.67583581386134, 0.0580954134929925, 0.307517499197274, 0.0603509526699781, 0.636024392675608, 0.753542394842952, 0.21510361880064, 0.934632363496348, 0.496509767370299, 0.826519216643646, 0.192319612484425, 0.722775686532259, 0.624518541619182, 0.397913841763511, 0.621281221276149, 0.199609044240788, 0.440858467947692, 0.272119887406006, 0.465551929082721, 0.0529092361684889, 0.0583762903697789, 0.715580987045541, 0.555090260691941, 0.219055795110762, 0.217679189983755, 0.0881865932606161, 0.738201884087175, 0.780529346084222, 0.736180980689824, 0.312460485845804, 0.939190223347396, 0.711132914293557, 0.186645412817597, 0.793599577853456, 0.102208665572107, 0.912131366552785, 0.155149355763569, 0.804908289341256, 0.402236063731834, 0.756633437704295, 0.333403246710077, 0.314398666610941, 0.687281941762194, 0.770855955081061, 0.0587603733874857, 0.721087303478271, 0.721969930687919, 0.676580390194431, 0.537996517727152, 0.585001531522721, 0.163080635480583, 0.402489589527249, 0.464902135077864, 0.899498186772689, 0.284756569191813, 0.969014471163973, 0.670498718973249, 0.0876998724415898, 0.154005118645728, 0.647283322410658, 0.129795537795871, 0.155485160183161, 0.633790266700089, 0.319980609696358, 0.648528591962531, 0.594667651457712, 0.434102111263201, 0.594436356332153, 0.50658523873426, 0.758148146793246, 0.363382170209661, 0.107449307804927, 0.454093721928075, 0.943232609191909, 0.359915254404768, 0.344544864725322, 0.785293504362926, 0.295864536426961, 0.871337817050517, 0.397127327742055, 0.482310571242124, 0.628229327267036, 0.417734856950119, 0.260231292108074, 0.230754937045276, 0.174125633900985, 0.783501997822896, 0.292564196279272, 0.160698799882084, 0.413583043962717, 0.856007620925084, 0.777352550532669, 0.0463405265472829, 0.561227935599163, 0.567409682320431, 0.204327141400427, 0.888142348034307, 0.138849928742275, 0.500511499820277, 0.3347853270825, 0.402017532847822, 0.146048160502687, 0.189671942032874, 0.160793158225715, 0.150622476125136, 0.0154782456811517, 0.816745734540746, 0.297987798927352, 0.804792273323983, 0.0681433277204633, 0.436249838210642, 0.416054096771404, 0.999760680133477, 0.731322951382026, 0.0543922376818955, 0.602873025462031, 0.410172333475202, 0.540288311894983, 0.860143091296777, 0.065358345862478, 0.0999145430978388, 0.194351799087599, 0.430188124300912, 0.776684749638662, 0.991646197391674, 0.7389815824572, 0.700902419164777, 0.0345383337698877, 0.624745306093246, 0.423928930191323, 0.566342561040074, 0.935805591056123, 0.731048145797104, 0.934402227867395, 0.0447708002757281, 0.261931914603338, 0.6693906779401, 0.327200062340125, 0.320426895283163, 0.180560087785125, 0.293011327506974, 0.269481901777908, 0.372360857203603, 0.62846022285521, 0.109363480703905, 0.0533628673292696, 0.338058026274666, 0.250407379353419, 0.785851757274941, 0.327291144058108, 0.80170462350361, 0.521838063374162, 0.541126708965749, 0.0667235280852765, 0.0203435954172164, 0.502500868868083, 0.979274860117584, 0.5072950522881, 0.960687681101263, 0.479987863684073, 0.611070688581094, 0.0514068321790546, 0.971503060311079, 0.375734233530238, 0.247943464899436, 0.575537464348599, 0.0390538249630481, 0.733416815986857, 0.0778198684565723, 0.418188250623643, 0.116635719547048, 0.689199139131233, 0.485791696002707, 0.096341093769297, 0.139355861581862, 0.587064001243562, 0.418319360818714, 0.0274540043901652, 0.547020348254591, 0.797134490683675, 0.844900883268565, 0.878858509706333, 0.462114252848551, 0.773322874680161, 0.0864636998157948, 0.961529123596847, 0.882507271366194, 0.172536838334054, 0.934636953752488, 0.178008367307484, 0.716182316187769, 0.0423396213445812, 0.936804292490706, 0.22156524262391, 0.752464625518769, 0.960585891269147, 0.439000085927546, 0.0665116147138178, 0.906883512856439, 0.374464111868292, 0.484488651156425, 0.393207895569503, 0.712899197591469, 0.129024474415928, 0.339749775826931, 0.231781291542575, 0.154479980934411, 0.903795845108107, 0.989031019853428, 0.746162181254476, 0.213015564251691, 0.415993050904945, 0.571352383354679, 0.778056978713721, 0.00915078050456941, 0.570271477336064, 0.0317962716799229, 0.218463117256761, 0.274315337650478, 0.886024556588382, 0.709696561330929, 0.461724564898759, 0.430417675990611, 0.363288186257705, 0.66591161605902, 0.566240714397281, 0.999646010808647, 0.0111901476047933, 0.749636032618582, 0.46956133027561, 0.804516098229215, 0.455813683336601, 0.852729624835774, 0.45803223666735, 0.845051544019952, 0.291537370765582, 0.657602560240775, 0.67040081648156, 0.174868912203237, 0.0374169801361859, 0.266836009453982, 0.516287934733555, 0.790307165123522, 0.928810972953215, 0.252101328223944, 0.214595885248855, 0.8682898117695, 0.568287213332951, 0.90228201658465, 0.598694844171405, 0.625393508933485, 0.0661655433941633, 0.721393822459504, 0.0891210522968322, 0.135238975752145, 0.786554198013619, 0.122001944808289, 0.720206668367609, 0.259706438751891, 0.997457416262478, 0.0898134200833738, 0.824357136851177, 0.368677517632023, 0.196810930268839, 0.92273959540762, 0.681981563568115, 0.151233068434522, 0.415292842546478, 0.963477604556829, 0.253899904899299, 0.572757981019095, 0.63679999532178, 0.221597601892427, 0.991757364477962, 0.908624045317993, 0.657306246925145, 0.676337619544938, 0.629001817200333, 0.0133251717779785, 0.88896770728752, 0.866533389082178, 0.174317252123728, 0.60024552186951, 0.945180559763685, 0.293443427188322, 0.848851165734231, 0.821596736088395, 0.134604095481336, 0.301887689856812, 0.535900844959542, 0.686526627512649, 0.268304632278159, 0.0512225835118443, 0.823661993257701, 0.441393199842423, 0.718133981572464, 0.344217163743451, 0.64072206011042, 0.505394053645432, 0.609942297451198, 0.898664819076657, 0.297089596744627, 0.747139846906066, 0.581321403384209, 0.24603083007969, 0.982462165877223, 0.853730100672692, 0.494244812522084, 0.978896883781999, 0.167862614849582, 0.787976473104209, 0.0252467994578183, 0.632478474639356, 0.500792913604528, 0.41133521636948, 0.133170369081199, 0.0470816020388156, 0.623587123351172, 0.209150786278769, 0.0641590154264122, 0.797946457751095, 0.0298279742710292, 0.914231777423993, 0.659854874480516, 0.704973546322435, 0.869917091215029, 0.715344665106386, 0.549590564565733, 0.137277888134122, 0.976889185374603, 0.626279140124097, 0.961319485679269, 0.0831998495850712, 0.922029700363055, 0.900273577542976, 0.612139881821349, 0.339995191665366, 0.0952537469565868, 0.0784512520767748, 0.433149905642495, 0.674015817465261, 0.607712703291327, 0.953782380791381, 0.159560185158625, 0.491192035609856, 0.807453780667856, 0.435878161573783, 0.5318827428855, 0.846487403847277, 0.640469554811716, 0.96163545595482, 0.573677884880453, 0.47628186433576, 0.866199502488598, 0.801398084266111, 0.796074990415946, 0.444908343954012, 0.867319835349917, 0.653335978277028, 0.495342239737511, 0.785706196911633, 0.172105418518186, 0.9421558144968, 0.365360734052956, 0.733946435619146, 0.943118458613753, 0.612571882782504, 0.212003751192242, 0.810083190444857, 0.182777824811637, 0.798180683050305, 0.00920043350197375, 0.494716780493036, 0.687438399996608, 0.0527092409320176, 0.197720135794953, 0.2176235136576, 0.243791445158422, 0.673402638640255, 0.230539979645982, 0.468835080740973, 0.156861631432548, 0.664493630640209, 0.397076524794102, 0.087813209509477, 0.344832311850041, 0.384195354534313, 0.660692357225344, 0.339107826352119, 0.927357435692102, 0.358971912181005, 0.219133164267987, 0.478341655340046, 0.594158596592024, 0.554149187169969, 0.384927607839927, 0.865034208865836, 0.307734568603337, 0.0254492782987654, 0.591589311137795, 0.703954651718959, 0.761199845699593, 0.862660236423835, 0.06697480333969, 0.475928724743426, 0.614812973188236, 0.468086575856432, 0.864138919860125, 0.897483170032501, 0.701602602144703, 0.0775255314074457, 0.707481443881989, 0.798579805064946, 0.976799877826124, 0.000255756545811892, 0.618907721480355, 0.276702802162617, 0.998608351685107, 0.35402343259193, 0.140340921934694, 0.392886673100293, 0.308987928554416, 0.293298616306856, 0.739105033921078, 0.478189148008823, 0.803306907182559, 0.967936478788033, 0.778631715569645, 0.972440607147291), actual = c(0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.685392544604838, 0.552945346571505, 0.0549710616469383, 0.277156794909388, 0.928404515376315, 0.442391449352726, 0.809661687817425, 0.385785371996462, 0.21236902102828, 0.849679732695222, 0.835597373778, 0.701369449030608, 0.738650265382603, 0.409968496533111, 0.408511461690068, 0.108972945949063, 0.999398496001959, 0.18613567808643, 0.717582976212725, 0.633615267230198, 0.490025891223922, 0.5431745541282, 0.418777758255601, 0.624445757130161, 0.223468976793811, 0.452905878191814, 0.973770106211305, 0.288306328002363, 0.0592717560939491, 0.786002967739478, 0.402153306407854, 0.125431847060099, 0.88317692023702, 0.921581645961851, 0.964478862239048, 0.845216770889238, 0.978768471861258, 0.836407164810225, 0.289165403461084, 0.206571588059887, 0.582869115518406, 0.103583892341703, 0.140415429836139, 0.609560237964615, 0.901483929716051, 0.726319778477773, 0.194152450188994, 0.0995915501844138, 0.195901321945712, 0.575820997124538, 0.975599021650851, 0.22082484816201, 0.0304569434374571, 0.895330210914835, 0.223010004963726, 0.047729337355122, 0.133136984426528, 0.492333118338138, 0.915886246366426, 0.640790815465152, 0.131165718194097, 0.633892190642655, 0.142986416583881, 0.937603765400127, 0.791374916443601, 0.834497923031449, 0.49561577104032, 0.528267652262002, 0.775853801751509, 0.993234969675541, 0.905420055380091, 0.132965714437887, 0.886829272843897, 0.515789920464158, 0.820075285620987, 0.68001894466579, 0.212058026343584, 0.606568024028093, 0.794661798747256, 0.828409123700112, 0.67275130841881, 0.898688039975241, 0.131386864930391, 0.897141864523292, 0.151626389939338, 0.374097239924595, 0.618819236056879, 0.876688191667199, 0.461616212269291, 0.972585056675598, 0.227631727233529, 0.0195151478983462, 0.336391987511888, 0.408430168870836, 0.674893293529749, 0.99252058332786, 0.909850798081607, 0.168122111586854, 0.328638338018209, 0.145926653407514, 0.220493923407048, 0.792094809701666, 0.82430913252756, 0.33840174949728, 0.259677535621449, 0.234647330129519, 0.320452877786011, 0.386816688813269, 0.0627969044726342, 0.630604192847386, 0.165115367621183, 0.327905804850161, 0.542560923146084, 0.245850306935608, 0.396903497166932, 0.0178266288712621, 0.870448786066845, 0.396384010789916, 0.0412149522453547, 0.334942889166996, 0.329032262787223, 0.474811322288588, 0.243183072889224, 0.451876106206328, 0.874328330159187, 0.763631864916533, 0.106072408379987, 0.895402636844665, 0.101339197950438, 0.864651352632791, 0.863117389846593, 0.6914319077041, 0.698631244944409, 0.422510554548353, 0.093311307951808, 0.261563312495127, 0.501127529190853, 0.900848506717011, 0.464468300109729, 0.698329682694748, 0.554046293254942, 0.795319905271754, 0.0363902647513896, 0.892017326550558, 0.553290354786441, 0.998985794838518, 0.504514672560617, 0.592002763645723, 0.460850958712399, 0.549310221802443, 0.349130524788052, 0.914465004345402, 0.577041003387421, 0.0602730002719909, 0.145299986703321, 0.13699941476807, 0.181141960434616, 0.130476458230987, 0.293051895918325, 0.927165996516123, 0.15581330889836, 0.0169127753470093, 0.289862253936008, 0.22482736222446, 0.602120436728001, 0.852759415050969, 0.227732076542452, 0.132013008696958, 0.261092422995716, 0.0010515539906919, 0.721423434792086, 0.382889855653048, 0.831707053352147, 0.931557612260804, 0.972094179829583, 0.67170403059572, 0.695349431363866, 0.340532445115969, 0.0965159989427775, 0.764603218296543, 0.547960695810616, 0.508130946429446, 0.362124183680862, 0.80682912399061, 0.955387118505314, 0.664234777214006, 0.253106993157417, 0.623223192058504, 0.438953099539503, 0.338722909800708, 0.744315329706296, 0.408866625744849, 0.113020514370874, 0.462793815648183, 0.0494428516831249, 0.842014280846342, 0.486108079785481, 0.808982119662687, 0.512118939543143, 0.539894368499517, 0.0343016479164362, 0.978016432141885, 0.277483282145113, 0.178247381001711, 0.84892401448451, 0.0196771179325879, 0.462351440219209, 0.388323849532753, 0.581752539379522, 0.117398872505873, 0.0495927408337593, 0.127747579244897, 0.263043291866779, 0.759500283515081, 0.474474696908146, 0.125239616725594, 0.489448970416561, 0.130865140585229, 0.11430700100027, 0.945496505126357, 0.784386151004583, 0.820432963781059, 0.879104811465368, 0.0667279770132154, 0.721715616993606, 0.87583261472173, 0.555486432975158, 0.583717046072707, 0.767044677166268, 0.623199820518494, 0.96488294377923, 0.692541708936915, 0.727404096629471, 0.258140116464347, 0.424215697450563, 0.110541159752756, 0.739347286755219, 0.533685798058286, 0.52762027643621, 0.72280322574079, 0.79497513663955, 0.486248600529507, 0.180189822800457, 0.979772293707356, 0.563932994613424, 0.758829011116177, 0.0494193928316236, 0.688059498555958, 0.9527697567828, 0.286115628201514, 0.675470180111006, 0.693016093922779, 0.949775030137971, 0.907270362135023, 0.116777439601719, 0.595099963713437, 0.226835506269708, 0.899807789828628, 0.390767010161653, 0.208064819918945, 0.294793486362323, 0.735690301982686, 0.00586177431978285, 0.612523305462673, 0.570233853068203, 0.284538016887382, 0.792669548885897, 0.146594065474346, 0.488494092598557, 0.290167115395889, 0.678943697130308, 0.699363519437611, 0.231949687935412, 0.520560967735946, 0.436166669474915, 0.777987027540803, 0.726546343183145, 0.94085220200941, 0.552061106776819, 0.843579007778317, 0.54785910830833, 0.327646553982049, 0.252475389512256, 0.28552475082688, 0.0474177731666714, 0.657731596380472, 0.798041698988527, 0.474975392455235, 0.138796037063003, 0.301129904109985, 0.491699517471716, 0.515549366362393, 0.273149782326072, 0.946522040292621, 0.862875586375594, 0.233918213052675, 0.553483890369534, 0.821558035211638, 0.541925408644602, 0.775537198875099, 0.272007958497852, 0.878196256933734, 0.76406924542971, 0.686920221196488, 0.0401044215541333, 0.502760162809864, 0.978872907347977, 0.939871242968366, 0.708612972870469, 0.370100145926699, 0.630258092889562, 0.954215539619327, 0.830736024770886, 0.550052553648129, 0.908630111953244, 0.476253726519644, 0.233791101258248, 0.94201593589969, 0.503395732259378, 0.858415513997898, 0.602250969968736, 0.632421027170494, 0.425487031694502, 0.468061820836738, 0.219079440459609, 0.459414745215327, 0.262112104333937, 0.412328769918531, 0.313032211968675, 0.13726058206521, 0.244047795888036, 0.160341735230759, 0.18683666898869, 0.00966710434295237, 0.0767464602831751, 0.672939686337486, 0.871093056164682, 0.23815528396517, 0.82099283253774, 0.461208489723504, 0.495225835591555, 0.487277734326199, 0.603643057402223, 0.611444548936561, 0.257947106380016, 0.224493000423536, 0.885258088121191, 0.120069249765947, 0.765270349569619, 0.197166349971667, 0.11257397942245, 0.197303640423343, 0.144179578404874, 0.867299795616418, 0.764421218074858, 0.244497198378667, 0.596941454568878, 0.872234480921179, 0.933030586922541, 0.0168767704162747, 0.896006307797506, 0.218291143886745, 0.0638532952871174, 0.540281687863171, 0.37620681617409, 0.722690478898585, 0.962901455583051, 0.391442980384454, 0.575737302657217, 0.676936917705461, 0.377739791525528, 0.161396618001163, 0.395740824751556, 0.113412403967232, 0.296445636311546, 0.469108166871592, 0.594138155225664, 0.454292956972495, 0.439009610330686, 0.87355143413879, 0.0279606964904815, 0.671547478064895, 0.708920197561383, 0.609761438565329, 0.935960764531046, 0.606445811456069, 0.826835321029648, 0.452537759207189, 0.58078020485118, 0.691088185878471, 0.149302474223077, 0.349849137710407, 0.782006917055696, 0.0448976543266326, 0.517616762081161, 0.127123750047758, 0.424303924432024, 0.322364009683952, 0.673165929038078, 0.03181436448358, 0.69846605672501, 0.380816082004458, 0.114047834882513, 0.0878828286658973, 0.136155099142343, 0.950333058834076, 0.170314725488424, 0.0844552940689027, 0.847448803950101, 0.109363226452842, 0.301740119233727, 0.436034506652504, 0.676067718071863, 0.261093886569142, 0.992101723561063, 0.0129081124905497, 0.588857400231063, 0.865946030244231, 0.644568896386772, 0.789653083076701, 0.220977502176538, 0.71586421946995, 0.710691822692752, 0.613677001092583, 0.76398889394477, 0.817923079244792, 0.544171576621011, 0.467423394322395, 0.93269285489805, 0.913948382018134, 0.511464489158243, 0.997006886638701, 0.965521258069202, 0.486846007173881, 0.545215387362987, 0.624617386609316, 0.741889812517911, 0.54299485986121, 0.360945383552462, 0.527571034152061, 0.475201964378357, 0.891661800211295, 0.123210598016158, 0.7813656530343, 0.858606266323477, 0.66199417039752, 0.259092319291085, 0.536757592810318, 0.728866626275703, 0.717952410690486, 0.282518431544304, 0.395060200477019, 0.611194197786972, 0.645288466941565, 0.84487086138688, 0.583842790918425, 0.325119317974895, 0.548019734909758, 0.529546447098255, 0.265338315628469, 0.114922933280468, 0.510472171707079, 0.341881268890575, 0.386939833406359, 0.326980518642813, 0.135722417850047, 0.406610344769433, 0.80232766433619, 0.285633325343952, 0.380112708546221, 0.341136700706556, 0.0905510624870658, 0.762493859743699, 0.4608506008517, 0.76947141578421, 0.544387266039848, 0.453090735245496, 0.794568324461579, 0.657896166667342, 0.43670529150404, 0.586217844625935, 0.404871229082346, 0.843507368117571, 0.786437439266592, 0.859742407221347, 0.76229827827774, 0.517887582769617, 0.558092205785215, 0.58438008883968, 0.942224896047264, 0.686051026219502, 0.569714234676212, 0.266571012092754, 0.034740992821753, 0.743092760909349, 0.969057478243485, 0.753294005757198, 0.049311499344185, 0.0765719672199339, 0.469639787916094, 0.0945355931762606, 0.824237670982257, 0.649563867598772, 0.759942336939275, 0.101489917142317, 0.85333422315307, 0.156221463810652, 0.892012549797073, 0.924473420251161, 0.682449207874015, 0.0204220118466765, 0.55608221841976, 0.190523670287803, 0.934664654079825, 0.305707597406581, 0.347478651208803, 0.103279943345115, 0.399169348878786, 0.511201595654711, 0.697449519997463, 0.0405569383874536, 0.842017962597311, 0.190939564490691, 0.603848763974383, 0.444131220923737, 0.833602997008711, 0.147726388880983, 0.496033394709229, 0.700503806117922, 0.697952098911628, 0.522938688052818, 0.0520463197026402, 0.125065928325057, 0.0899186376482248, 0.764946425333619, 0.930796399246901, 0.53132253116928, 0.107155036646873, 0.0564404006581753, 0.491367225768045, 0.575398423243314, 0.997590450337157, 0.605960440589115, 0.880155782913789, 0.145143626723439, 0.700839675730094, 0.656077375402674, 0.645213802810758, 0.601159092970192, 0.79744310490787, 0.894689760403708, 0.987157503142953, 0.469338095746934, 0.334072562633082, 0.794788723113015, 0.0873208488337696, 0.924507250776514, 0.477980375057086, 0.654412667267025, 0.0354052307084203, 0.785269080195576, 0.00207874504849315, 0.699593836907297, 0.446912803919986, 0.886213042540476, 0.821033768123016, 0.392429709900171, 0.466371219605207, 0.223841377301142, 0.625728394836187, 0.758228118764237, 0.841921673621982, 0.117894894676283, 0.892789121018723, 0.784612087532878, 0.351296911714599, 0.813567207194865, 0.0970613029785454, 0.428016298916191, 0.402100303443149, 0.0118133646901697, 0.997669242788106, 0.865237609483302, 0.0844968687742949, 0.634991681668907, 0.864356623729691, 0.0734616161789745, 0.739698074292392, 0.626657837769017, 0.673178381286561, 0.212066549109295, 0.0308462886605412, 0.825990684796125, 0.624716044869274, 0.489719494711608, 0.311332343611866, 0.728652058634907, 0.361467853421345, 0.513618069933727, 0.549979403149337, 0.679433570476249, 0.505115907639265, 0.718301683897153, 0.36087993811816, 0.0998931631911546, 0.841057005338371, 0.692098072497174, 0.437440197914839, 0.979614012409002, 0.811182550387457, 0.182613316457719, 0.708728715544567, 0.599088271846995, 0.00997188175097108, 0.185404577758163, 0.934520876267925, 0.0597665412351489, 0.693459834204987, 0.233618439175189, 0.561976995086297, 0.656143271597102, 0.402995225274935, 0.231898864964023, 0.385200057644397, 0.84011377603747, 0.302048044279218, 0.820770106045529, 0.137230853317305, 0.0959851914085448, 0.893797992961481, 0.40089476807043, 0.16017623571679, 0.823142660781741, 0.963844797806814, 0.826045476365834, 0.202052644453943, 0.0534526687115431, 0.33764166617766, 0.791500850580633, 0.0126045129727572, 0.106777528300881, 0.511197529034689, 0.447712708963081, 0.488691903185099, 0.805269968230277, 0.330260137561709, 0.00403273361735046, 0.682738100644201, 0.263056251686066, 0.0805696991737932, 0.960408053826541, 0.98691148054786, 0.11864525382407, 0.2280210533645, 0.0661998719442636, 0.124735639896244, 0.012051151599735, 0.342669898411259, 0.797163086012006, 0.929518600460142, 0.78551756683737, 0.723806326044723, 0.0680929096415639, 0.333418184425682, 0.604049985995516, 0.588901999173686, 0.158785399748012, 0.773094805888832, 0.577602349454537, 0.122221488971263, 0.821194801246747, 0.213549463776872, 0.728308262536302, 0.307477179216221, 0.0608239572029561, 0.777975848177448, 0.246306333690882, 0.31102109584026, 0.532541509717703, 0.69230522098951, 0.463723189663142, 0.204013393493369, 0.742876348318532, 0.777475623879582, 0.0154320325236768, 0.905592341907322, 0.753669902682304, 0.379223886877298, 0.0194292364176363, 0.149449314223602, 0.514333897503093, 0.48723271372728, 0.457963325781748, 0.0146865365095437, 0.392326760105789, 0.161100027384236, 0.263795301318169, 0.797416667686775, 0.595795199507847, 0.0409927137661725, 0.380132269347087, 0.0825431139674038, 0.318601910257712, 0.823136059101671, 0.442699896870181, 0.0553729136008769, 0.15197766199708, 0.568675088696182, 0.691520801046863, 0.920679031405598, 0.949295777594671, 0.0903213259298354, 0.530930964741856, 0.690288411919028, 0.131450584391132, 0.499330263119191, 0.48767996695824, 0.719915413064882, 0.719050222076476, 0.650004095863551, 0.376619879622012, 0.692440996179357, 0.682803335366771, 0.135332313133404, 0.863641607575119, 0.522829128662124, 0.88491936121136, 0.947835130151361, 0.887154090218246, 0.804258592659608, 0.0731468205340207, 0.58056813548319, 0.178625576663762, 0.691261882893741, 0.375272534321994, 0.712316933320835, 0.368577402783558, 0.461334206163883, 0.381144154118374, 0.672563599888235, 0.742187357973307, 0.872748831985518, 0.931976725580171, 0.127687867730856, 0.846379606984556, 0.429424325702712, 0.0132857402786613, 0.807020167121664, 0.473400526447222, 0.415194514207542, 0.386908970074728, 0.251837532734498, 0.196707183029503, 0.505027221748605, 0.111788414884359, 0.913498013513163, 0.697987970197573, 0.834545511053875, 0.485003784531727, 0.515294553013518, 0.0989260436035693, 0.353443868691102, 0.709242891287431, 0.0628008698113263, 0.166506160050631, 0.325793029507622, 0.313469323562458, 0.0780850274022669, 0.729257140075788, 0.978296641726047, 0.757571453228593, 0.965078906621784, 0.369563590502366, 0.563217077404261, 0.811091692419723, 0.41408561100252, 0.230269028572366, 0.919196429662406, 0.806559909833595, 0.868422487052158, 0.801440900424495, 0.952773401513696, 0.724782135337591, 0.978564077289775, 0.769473444670439, 0.504154195077717, 0.576247555436566, 0.448042633477598, 0.862028595991433, 0.351471352390945, 0.111677842913195, 0.403640428092331, 0.894497214118019, 0.650692212162539, 0.658094264334068, 0.393504303647205, 0.914203024469316, 0.683265777537599, 0.130158931016922, 0.306470516836271, 0.0613527144305408, 0.579838539706543, 0.251089996192604, 0.7785794720985, 0.982400462497026, 0.149255797965452, 0.378315642476082, 0.495263096177951, 0.432003122987226, 0.269167070975527, 0.674729407066479, 0.621482167160138, 0.835951894288883, 0.709705047309399, 0.617963016033173, 0.786364805186167, 0.995243287179619, 0.664532497059554, 0.721153575228527, 0.663762756390497, 0.316391216823831, 0.115899347001687, 0.10225121001713, 0.038373181829229, 0.35811400343664, 0.285119784763083, 0.842042356031016, 0.594562938204035, 0.994114590110257, 0.642365223262459, 0.204888046020642, 0.401398086221889, 0.76480012992397, 0.935193688841537, 0.686783637851477, 0.507623140932992, 0.0805382847320288, 0.0619663905818015, 0.667461581295356, 0.728783428901806, 0.682195336557925, 0.295345672173426, 0.157694715075195, 0.213646512944251, 0.476868513738737, 0.0971132551785558, 0.256724345032126, 0.15718304598704, 0.525999184232205, 0.337987888371572, 0.379393334733322, 0.305606876499951, 0.550560479983687, 0.830649112351239, 0.201816395390779, 0.399712543701753, 0.742868107045069, 0.398174594854936, 0.0783868855796754, 0.194220694014803, 0.448239114833996, 0.532251112395898, 0.606982329627499, 0.335599784506485, 0.547137459041551, 0.760661824373528, 0.211680561536923, 0.619685814483091, 0.576926803681999, 0.351056125946343, 0.480556381633505, 0.737523016519845, 0.144960356876254, 0.537179647712037, 0.197159144328907, 0.840903687523678, 0.938603275921196, 0.754863422363997, 0.121869378490373, 0.991630584700033, 0.999275462003425, 0.763773341896012, 0.480255149072036, 0.186914305435494, 0.868547147139907, 0.514782841783017, 0.216779651818797, 0.42167773284018, 0.824483406729996, 0.630640964955091, 0.564664202742279, 0.718389814952388, 0.860483513679355, 0.476052556885406, 0.55554520781152, 0.755128781311214, 0.968026128364727, 0.342482621083036, 0.151432823622599, 0.0569588060025126, 0.784527393057942, 0.675286122830585, 0.228519264608622, 0.55895495833829, 0.504684674786404, 0.877419323660433, 0.125734829809517, 0.40827786992304, 0.631302915746346, 0.233122523408383, 0.0424438761547208, 0.875992482993752, 0.834848527796566, 0.120137332007289, 0.994336056057364, 0.948080490110442, 0.93054562411271, 0.950262861093506, 0.897039752686396, 0.292342039523646, 0.475991006940603, 0.0372487574350089, 0.814195572631434, 0.0880190485622734, 0.44163209781982, 0.947649454465136, 0.858809467405081, 0.887395390076563, 0.0269970237277448, 0.927496203687042, 0.0969903827644885, 0.706980552524328, 0.466311846626922, 0.887593663064763, 0.551627945154905, 0.0114153276663274, 0.558452856028453, 0.132089166436344, 0.42068656720221, 0.100531770382077, 0.811193660600111, 0.120117494137958, 0.368523018667474, 0.248431279789656, 0.0771133294329047, 0.560164335882291, 0.883080220548436, 0.849116239929572, 0.12729654321447, 0.0766493247356266, 0.626020824769512, 0.207785147242248, 0.988726068753749, 0.156635652529076, 0.622317708097398, 0.0165878320112824, 0.935980363981798, 0.692497100913897, 0.131601270986721, 0.505891834385693, 0.721050393069163, 0.332069699652493, 0.649855670519173, 0.119766145711765, 0.652850409504026, 0.380341894226149, 0.178952842950821, 0.665656587574631, 0.279113310622051, 0.431210316950455, 0.76195728010498, 0.166969639714807, 0.614204965764657, 0.547232748474926, 0.361426238436252, 0.261052442947403, 0.824771769111976, 0.329246178502217, 0.801823309622705, 0.146689883433282, 0.333501640474424, 0.791862331796438, 0.0154840892646462, 0.0710157498251647, 0.464704452548176, 0.267924319719896, 0.770210889866576, 0.153138101333752, 0.267505802912638, 0.554315475746989, 0.427390168653801, 0.776378612732515, 0.624287484912202, 0.113474270096049, 0.462794654536992, 0.942624387796968, 0.161274017533287, 0.443786200601608, 0.20260110241361, 0.0215767021290958), actual = c(0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"