Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>       0.00000000       1.00000000       0.02641662       0.76707961 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>       0.50000000       0.48929198       0.48929198       0.51837552 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.870685565052554, 0.825679440516979, 0.791529055917636, 0.905409732367843, 0.579904874321073, 0.565693333512172, 0.0839445719029754, 0.812055721413344, 0.738656066125259, 0.361860551871359, 0.208047647261992, 0.763833238510415, 0.594061199808493, 0.490816919365898, 0.105306081473827, 0.389898097142577, 0.604768374469131, 0.598621253855526, 0.497324022930115, 0.745365398703143, 0.296222659759223, 0.43174015567638, 0.146852807840332, 0.403394584078342, 0.86324148462154, 0.0521838294807822, 0.389426870737225, 0.316047718049958, 0.240593102294952, 0.744995172368363, 0.473190421471372, 0.28727606870234, 0.105436899000779, 0.163719225209206, 0.545984072145075, 0.614652320509776, 0.323165501467884, 0.547131319064647, 0.827862143050879, 0.313038113061339, 0.822555967140943, 0.483704398619011, 0.892567297210917, 0.151298328768462, 0.456919588847086, 0.996348984772339, 0.424366501625627, 0.978784451261163, 0.692430128343403, 0.481954077957198, 0.817673402605578, 0.477575472090393, 0.155498213367537, 0.426698269555345, 0.654444187413901, 0.763121323194355, 0.490749714896083, 0.855652108089998, 0.412133838748559, 0.510696081444621, 0.955182798206806, 0.862410044996068, 0.123918329831213, 0.78352818521671, 0.713347367243841, 0.0616966444067657, 0.729112594155595, 0.859464917331934, 0.333911380497739, 0.972140840953216, 0.147990049328655, 0.379422072088346, 0.636455065570772, 0.796177953481674, 0.774177917744964, 0.794680333696306, 0.282147925114259, 0.927857926115394, 0.479944301769137, 0.911787496646866, 0.0774040254764259, 0.10424183215946, 0.559753704583272, 0.958761215209961, 0.101867575198412, 0.613205874571577, 0.831730286125094, 0.99229027540423, 0.378424362977967, 0.566482236841694, 0.17615095176734, 0.706959387054667, 0.344661809271201, 0.00679886410944164, 0.413212794344872, 0.541163053596392, 0.834440430393443, 0.663513290230185, 0.390744465636089, 0.383242816431448, 0.682034585392103, 0.0716460179537535, 0.013321227626875, 0.764764109393582, 0.705404989654198, 0.892750549595803, 0.195318658603355, 0.475420431466773, 0.654235386988148, 0.939639786724001, 0.0349102744366974, 0.49412672361359, 0.490699468646199, 0.286032702075318, 0.265259510604665, 0.019408582476899, 0.991070564137772, 0.11958011589013, 0.126234757248312, 0.14906870177947, 0.0825467705726624, 0.523951124865562, 0.415618476690724, 0.694126832997426, 0.500292379874736, 0.634908139239997, 0.730591374216601, 0.429392335703596, 0.191914841998369, 0.325616253074259, 0.0559402094222605, 0.160022821743041, 0.388312690192834, 0.716422015801072, 0.657676580362022, 0.828189921798185, 0.790525078540668, 0.9995401725173, 0.150233254069462, 0.893900086637586, 0.642789282137528, 0.79108137707226, 0.992714742664248, 0.721501447958872, 0.210627031279728, 0.0580248413607478, 0.585979744559154, 0.567688539624214, 0.685156798223034, 0.493582411669195, 0.429751461138949, 0.100887929555029, 0.589713941095397, 0.236085338750854, 0.332845553290099, 0.406868670834228, 0.276348644169047, 0.693906988948584, 0.99896136415191, 0.0817066126037389, 0.748637873679399, 0.753471598029137, 0.189479244872928, 0.61358090583235, 0.00651113595813513, 0.822254670783877, 0.809837594628334, 0.945951311849058, 0.157764774048701, 0.342486657667905, 0.407300926279277, 0.698399605695158, 0.84134884015657, 0.537372286897153, 0.597546858945861, 0.974021247820929, 0.518512868555263, 0.532777383225039, 0.164801762904972, 0.348500952124596, 0.72661302681081, 0.931342726806179, 0.901910898042843, 0.78424692270346, 0.301380621036515, 0.0202746589202434, 0.500928326277062, 0.610543456859887, 0.66599114285782, 0.18825081945397, 0.707696620840579, 0.71143493312411, 0.657045417232439, 0.204335643677041, 0.438798257848248, 0.11118725198321, 0.893881137482822, 0.901684110052884, 0.554265895625576, 0.951859139837325, 0.542636164929718, 0.654830881161615, 0.816169285681099, 0.36576019693166, 0.154737807810307, 0.738308881176636, 0.235612316522747, 0.970774879911914, 0.963958522537723, 0.15451245335862, 0.961563976015896, 0.744968814542517, 0.373729061335325, 0.332643995061517, 0.655075672548264, 0.799123853212222, 0.974373102653772, 0.268393308622763, 0.17233007051982, 0.485043255845085, 0.226789605803788, 0.766602206742391, 0.586287553422153, 0.726826560450718, 0.317357066785917, 0.885256365174428, 0.315037995576859, 0.968822266208008, 0.879859169013798, 0.205821727169678, 0.409243779489771, 0.365909684682265, 0.569845498772338, 0.360571931349114, 0.24838273669593, 0.522307951236144, 0.187890723580495, 0.958693272201344, 0.656395010184497, 0.683044717647135, 0.912833848502487, 0.914680641843006, 0.0365551393479109, 0.0421538795344532, 0.0995568532962352, 0.941233511082828, 0.920227893162519, 0.356038019061089, 0.306881742784753, 0.698781423969194, 0.776418019086123, 0.447377217933536, 0.407901634462178, 0.478274347493425, 0.95669751171954, 0.177013797452673, 0.623363299993798, 0.891502732643858, 0.338251741835847, 0.773296369239688, 0.274171462282538, 0.304952838458121, 0.0780572907533497, 0.0843900868203491, 0.394651510054246, 0.410895477514714, 0.0945734546985477, 0.999341001268476, 0.0647733139339834, 0.48410557792522, 0.0674539003521204, 0.556278610602021, 0.536411919398233, 0.39802686101757, 0.177768202731386, 0.743569133337587, 0.993446483043954, 0.595076417317614, 0.0326860286295414, 0.031332855578512, 0.972217758418992, 0.0817527193576097, 0.76887634745799, 0.167785123223439, 0.392039123689756, 0.586997707141563, 0.301608463283628, 0.303072806214914, 0.639657749794424, 0.0565139227546751, 0.194621526170522, 0.474821417359635, 0.839712062152103, 0.844466693932191, 0.911042652558535, 0.198594542453066, 0.892846447648481, 0.166646269150078, 0.650907762348652, 0.895500637125224, 0.67653009109199, 0.292964323656633, 0.62194985547103, 0.638925054110587, 0.931333439657465, 0.477347693871707, 0.565365127520636, 0.643425199668854, 0.987370814196765, 0.9385315713007, 0.884706828277558, 0.460938752163202, 0.0820752170402557, 0.349083509994671, 0.0339113534428179, 0.700184557121247, 0.440692696720362, 0.676865349756554, 0.60921136662364, 0.757592057809234, 0.202782094012946, 0.473548439797014, 0.370970445219427, 0.19756498048082, 0.741497009526938, 0.270729867275804, 0.987797400681302, 0.108126881765202, 0.335159104550257, 0.596208738628775, 0.782352462178096, 0.0167370794806629, 0.111786239547655, 0.975862930295989, 0.229927345411852, 0.270166122587398, 0.510847578523681, 0.0606518005952239, 0.448046652134508, 0.577230588765815, 0.0241673900745809, 0.598095291526988, 0.444014432607219, 0.174226358067244, 0.700223522493616, 0.299163555260748, 0.417429046472535, 0.16488396585919, 0.58723752037622, 0.866341296583414, 0.885336796985939, 0.969660487957299, 0.00850833114236593, 0.18774191243574, 0.457017614971846, 0.507345526246354, 0.816742552444339, 0.284026263980195, 0.820930782472715, 0.0111823212355375, 0.546692271716893, 0.985622589243576, 0.725246933987364, 0.420226235641167, 0.911611526971683, 0.756626589922234, 0.785346440738067, 0.979394326219335, 0.270047892583534, 0.282989381346852, 0.990120687754825, 0.578799281734973, 0.20946834422648, 0.921665757428855, 0.642997437855229, 0.458400550996885, 0.901582185411826, 0.604456631932408, 0.21676377649419, 0.280741869006306, 0.412139443680644, 0.540143735939637, 0.0549211257603019, 0.331132841296494, 0.507327673723921, 0.0134074920788407, 0.436396062606946, 0.567335189785808, 0.724246744997799, 0.513024058192968, 0.47656805976294, 0.8503175645601, 0.500295285135508, 0.234011303633451, 0.657640770077705, 0.950506749097258, 0.677809338318184, 0.87271775607951, 0.969979657093063, 0.540439991280437, 0.53657585894689, 0.12523322715424, 0.169914283324033, 0.699369146022946, 0.570117929950356, 0.716736954869702, 0.810854813316837, 0.466735064284876, 0.827994495164603, 0.813683496788144, 0.297603098209947, 0.934426298830658, 0.536828421987593, 0.471483700210229, 0.919725155457854, 0.737750865984708, 0.903606528183445, 0.28518994525075, 0.555054811993614, 0.5330900689587, 0.73113040626049, 0.805067295907065, 0.63558076065965, 0.985243796603754, 0.118389263749123, 0.562219634652138, 0.595095607917756, 0.25601019943133, 0.309152998728678, 0.953691014554352, 0.176243272377178, 0.0644127062987536, 0.081287776818499, 0.388130201026797, 0.253172240452841, 0.129500166745856, 0.956648968858644, 0.777559124864638, 0.537445315392688, 0.888724330347031, 0.496254183351994, 0.77493275818415, 0.307203905424103, 0.901688514044508, 0.190993356518447, 0.360503105446696, 0.561382528394461, 0.0312561732716858, 0.672860424732789, 0.943844318855554, 0.324779243674129, 0.383356787031516, 0.402627832023427, 0.558784797089174, 0.0820093443617225, 0.526557703036815, 0.265146571444348, 0.060330499894917, 0.771442500408739, 0.562532238662243, 0.0982220103032887, 0.685124747687951, 0.729228832293302, 0.748087937943637, 0.541779384948313, 0.0844537813682109, 0.100393558619544, 0.0152290626429021, 0.12824802682735, 0.587725940393284, 0.667567645199597, 0.00625050044618547, 0.97633452503942, 0.0758394177537411, 0.710053255781531, 0.680357920238748, 0.245891196886078, 0.739889342337847, 0.571350895101205, 0.0729579923208803, 0.154620800167322, 0.575470627052709, 0.0637390371412039, 0.197988769272342, 0.33434924017638, 0.42719053174369, 0.552273642504588, 0.701128472574055, 0.343363165389746, 0.889201566111296, 0.0164986532181501, 0.0944286941085011, 0.469154613325372, 0.0840390573721379, 0.871590199647471, 0.494942009914666, 0.87318639899604, 0.142360503319651, 0.761602341895923, 0.312966487370431, 0.416375248460099, 0.892094872426242, 0.795045163482428, 0.303133434616029, 0.488380703609437, 0.450125407427549, 0.0198049533646554, 0.974438535282388, 0.97297262144275, 0.507133713457733, 0.219107031822205, 0.680275686318055, 0.149947815341875, 0.454922495409846, 0.642435984686017, 0.361523805418983, 0.834301446098834, 0.85708355368115, 0.0555613315664232, 0.668067738879472, 0.369966584490612, 0.642350381007418, 0.864628776209429, 0.273648010566831, 0.240501245250925, 0.712779068155214, 0.468466120306402, 0.0644726697355509, 0.990360487718135, 0.191971100633964, 0.822313733398914, 0.876793236471713, 0.659078607102856, 0.32811862626113, 0.119102918310091, 0.106259366264567, 0.117609485285357, 0.868645657086745, 0.338550613028929, 0.456517232116312, 0.189045518171042, 0.186257371213287, 0.91081783012487, 0.667080223560333, 0.441059855977073, 0.874811838148162, 0.695743484189734, 0.120126103516668, 0.403604450169951, 0.150374988559633, 0.0612704595550895, 0.127651568502188, 0.0492009655572474, 0.651328327599913, 0.968011133838445, 0.548114466713741, 0.139892962994054, 0.875498149776831, 0.112477748189121, 0.6698863056954, 0.573561598779634, 0.054482813924551, 0.0340490946546197, 0.793445465853438, 0.509752706158906, 0.818271657684818, 0.210071127861738, 0.847055320860818, 0.691615630872548, 0.591015067882836, 0.224358161212876, 0.832225944148377, 0.484584993682802, 0.375534811289981, 0.30496962973848, 0.925132547970861, 0.805107832653448, 0.298763034399599, 0.629753741901368, 0.239675035234541, 0.760609335033223, 0.321404114365578, 0.967398436274379, 0.231397754512727, 0.114245247561485, 0.807394895935431, 0.423506968189031, 0.509096902096644, 0.027247569989413, 0.779575859429315, 0.28883100207895, 0.902267100987956, 0.400255805347115, 0.0783127916511148, 0.531727969180793, 0.804099316010252, 0.882017097203061, 0.488064063945785, 0.0331720327958465, 0.0730729685164988, 0.383816910209134, 0.325818356126547, 0.407460403395817, 0.0556095463689417, 0.354396933456883, 0.911495561245829, 0.86822228692472, 0.435503093525767, 0.778145131655037, 0.232686081668362, 0.160946996416897, 0.318628983106464, 0.762316775042564, 0.813021956477314, 0.84323790948838, 0.593710997374728, 0.775143417296931, 0.621704935561866, 0.933329539606348, 0.822915874421597, 0.248901339946315, 0.864711890928447, 0.13783897086978, 0.10437565180473, 0.70420375559479, 0.307309218216687, 0.660471277311444, 0.621574186487123, 0.00463211024180055, 0.997169990325347, 0.18057090532966, 0.130022543016821, 0.0476676172111183, 0.138873912161216, 0.403121188981459, 0.409671318018809, 0.152628844371065, 0.302126542897895, 0.684943303000182, 0.225629882421345, 0.984172812663019, 0.893387989141047, 0.577184578869492, 0.0516627300530672, 0.0603124334011227, 0.0850531379692256, 0.219075996195897, 0.414866981096566, 0.475795897655189, 0.754651546245441, 0.131139660486951, 0.650303782429546, 0.702816493809223, 0.187439962755889, 0.381082865875214, 0.918537001125515, 0.614643061999232, 0.000373339280486107, 0.659360019955784, 0.868537087691948, 0.90912375645712, 0.51797184580937, 0.859000829048455, 0.866360970307142, 0.089238464133814, 0.647067294456065, 0.0400803661905229, 0.0661059389822185, 0.265148436417803, 0.270142826018855, 0.469162779627368, 0.511660071089864, 0.211440671700984, 0.630247551482171, 0.70293282228522, 0.490093236090615, 0.282363344915211, 0.816276147030294, 0.557125818217173, 0.346746006514877, 0.0484669399447739, 0.108652545372024, 0.200485041132197, 0.478827392915264, 0.710247019538656, 0.860233534593135, 0.621823159046471, 0.969109927071258, 0.164411348057911, 0.408549311570823, 0.689760126406327, 0.728861901676282, 0.395536666270345, 0.12586885667406, 0.826297221472487, 0.300740422913805, 0.662452877964824, 0.0555797663982958, 0.621640857541934, 0.652002373710275, 0.876184617634863, 0.132380952592939, 0.42190485005267, 0.181013311492279, 0.927185974782333, 0.591561312554404, 0.250041066203266, 0.934017291525379, 0.462635453324765, 0.998239557724446, 0.811092359712347, 0.525918595027179, 0.878176926635206, 0.427867598831654, 0.798964730696753, 0.11485438561067, 0.411457312526181, 0.698276808485389, 0.354388139210641, 0.765263418899849, 0.680708111496642, 0.929930191254243, 0.871798317646608, 0.0738488947972655, 0.639938612468541, 0.59268652019091, 0.61144564114511, 0.513243447290733, 0.809010043507442, 0.544581422582269, 0.651689295191318, 0.897351470775902, 0.974075923906639, 0.583140261936933, 0.255047225160524, 0.4196610909421, 0.279556204099208, 0.338888854952529, 0.248357992153615, 0.0378149508032948, 0.953883740352467, 0.262139498256147, 0.410378557164222, 0.97512721689418, 0.0475930508691818, 0.529988969676197, 0.848375043133274, 0.796615628758445, 0.600546062691137, 0.999940190464258, 0.763660466996953, 0.40386814950034, 0.825276396237314, 0.781480700941756, 0.842383805429563, 0.0973809731658548, 0.0588802569545805, 0.253441659035161, 0.833660116652027, 0.906934601720423, 0.714896761346608, 0.148278885986656, 0.771895575337112, 0.640977293020114, 0.568233630154282, 0.145783285610378, 0.0153667770791799, 0.260580414673313, 0.66052004438825, 0.880135622108355, 0.151966127334163, 0.31392879714258, 0.709374840604141, 0.622574158711359, 0.716624513966963, 0.134179311338812, 0.864830195438117, 0.827578622847795, 0.155239624902606, 0.92448033997789, 0.446490761591122, 0.0602100489195436, 0.911264010006562, 0.68187542213127, 0.389726673252881, 0.696082145208493, 0.888007351430133, 0.993336576968431, 0.47739026369527, 0.761886425549164, 0.167257172521204, 0.628856686409563, 0.63347489386797, 0.793639258714393, 0.779481788398698, 0.908861682750285, 0.773414320545271, 0.077593696070835, 0.999426684575155, 0.0312995449639857, 0.594152821460739, 0.452385649317876, 0.872715157922357, 0.333154903491959, 0.531948723364621, 0.461480552097782, 0.871800381923094, 0.610748420935124, 0.487941333791241, 0.266170707531273, 0.400079174432904, 0.194203659426421, 0.957353600068018, 0.192797393072397, 0.178400562610477, 0.802989748073742, 0.352169687859714, 0.271913475822657, 0.605543210869655, 0.910704814596102, 0.247866768622771, 0.00424090633168817, 0.423442900413647, 0.636777976294979, 0.894296265207231, 0.681090204045177, 0.544296716572717, 0.890208527212963, 0.045839692465961, 0.158377832733095, 0.544440683908761, 0.858695583650842, 0.324273506412283, 0.782563081709668, 0.708735628053546, 0.471419259672984, 0.3000620638486, 0.947354812873527, 0.08013836434111, 0.43726078607142, 0.621397994691506, 0.170285639585927, 0.0453389314934611, 0.168462338391691, 0.0993559509515762, 0.890051586786285, 0.530627586878836, 0.283947969321162, 0.332606346812099, 0.803382492624223, 0.90861847368069, 0.441768802003935, 0.295875964686275, 0.658095221733674, 0.793864964973181, 0.893958414206281, 0.645072288345546, 0.670446045929566, 0.305588230723515, 0.526637821225449, 0.541335713118315, 0.637346396921203, 0.715937728062272, 0.365173379424959, 0.704870481742546, 0.0903130602091551, 0.921945106470957, 0.291154983919114, 0.294386945432052, 0.45851780869998, 0.717306504491717, 0.890583898639306, 0.964453619439155, 0.0291050334926695, 0.1655101578217, 0.566489208489656, 0.465452697128057, 0.695437056012452, 0.310551362344995, 0.186906951013952, 0.688295704778284, 0.616971570532769, 0.915881401393563, 0.31191728473641, 0.730604725889862, 0.820889276452363, 0.179951900849119, 0.369960565119982, 0.100770184537396, 0.896573888370767, 0.169002074981108, 0.236714396392927, 0.736379242502153, 0.515241422457621, 0.381653052056208, 0.0888514805119485, 0.903291177935898, 0.519163193413988, 0.968485218705609, 0.132002547150478, 0.637347689131275, 0.465810327557847, 0.403127033030614, 0.430266257375479, 0.470344039378688, 0.252090567955747, 0.2467481426429, 0.120938997715712, 0.687740766908973, 0.521643781336024, 0.264364187140018, 0.40301325567998, 0.516983846202493, 0.977078213356435, 0.271252933656797, 0.18352007959038, 0.671573284547776, 0.0744734706822783, 0.0642449650913477, 0.447523512644693, 0.98052006168291, 0.893840664299205, 0.837698685703799, 0.791486715665087, 0.164828715147451, 0.735459137242287, 0.166448197327554, 0.598446825752035, 0.246405661571771, 0.884769191034138, 0.34510744875297, 0.666729659773409, 0.833183532115072, 0.991167073138058, 0.00539961596950889, 0.994126310572028, 0.879045582609251, 0.318153402535245, 0.820840498199686, 0.322412861511111, 0.44216394261457, 0.505788909038529, 0.648731511086226, 0.138711125124246, 0.585135580506176, 0.387172910850495, 0.749686426250264, 0.850897024385631, 0.892766420729458, 0.0198722071945667, 0.196628528181463, 0.257338400930166, 0.996173400897533, 0.201698781689629, 0.126958912471309, 0.126870560459793, 0.21008608629927, 0.175071423407644, 0.278672641143203, 0.0837347891647369, 0.242534185526893, 0.950940298149362, 0.967426400166005, 0.666776393773034, 0.787361162016168, 0.185530881630257, 0.845444017322734, 0.429976746439934, 0.592171314172447, 0.276999412570149, 0.680169695522636, 0.287521859398112, 0.242724641691893, 0.232124912319705, 0.617653084918857, 0.45500598475337, 0.380475832149386, 0.682021655607969, 0.780564144952223, 0.23561536683701, 0.693593131145462, 0.510551201179624, 0.0501802612561733, 0.7436393359676, 0.809534313157201, 0.179756284924224, 0.666593657340854, 0.0689010077621788, 0.632230149116367, 0.373749153222889, 0.486681019188836, 0.309185090009123, 0.784033652162179, 0.65052667632699, 0.728623712202534, 0.361403703456745, 0.665189442690462, 0.859779809601605, 0.233602387132123, 0.939871621550992), actual = c(1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.376647472847253, 0.0760711112525314, 0.50952438195236, 0.587455662433058, 0.853791499976069, 0.454679095419124, 0.103768871631473, 0.442512001376599, 0.678237279411405, 0.345376469194889, 0.640870698727667, 0.100852211704478, 0.902927802642807, 0.173111092299223, 0.377302757231519, 0.185493771685287, 0.942339658970013, 0.847167456289753, 0.203477394301444, 0.643179854843765, 0.302600342547521, 0.795381443807855, 0.23452171171084, 0.0608953705523163, 0.924903057515621, 0.164792448980734, 0.558542993618175, 0.929612438194454, 0.529581425711513, 0.650434549199417, 0.590078450506553, 0.906567773083225, 0.211920813890174, 0.632346404716372, 0.955674852943048, 0.646096821175888, 0.994604473467916, 0.152177882613614, 0.287792905699462, 0.588255571667105, 0.975788716226816, 0.717450661351904, 0.785610555438325, 0.734756303485483, 0.0691543354187161, 0.264651519246399, 0.417722071520984, 0.384518714854494, 0.0526461971458048, 0.327208166476339, 0.800324448384345, 0.243721749866381, 0.868676194921136, 0.61381359398365, 0.109398265369236, 0.728330647340044, 0.262507969047874, 0.0139752831310034, 0.782862316817045, 0.103925016475841, 0.621251057367772, 0.472773458342999, 0.9289468198549, 0.711638772627339, 0.0966273825615644, 0.218156822724268, 0.171579254092649, 0.140270812669769, 0.794114235555753, 0.0529037625528872, 0.899081237148494, 0.0471174772828817, 0.656837842194363, 0.890925204614177, 0.290220161434263, 0.359112739097327, 0.171396368648857, 0.40256708720699, 0.448814617935568, 0.431045328266919, 0.624196445103735, 0.231755299028009, 0.592973958700895, 0.551036321558058, 0.423210707725957, 0.429997163126245, 0.22545262100175, 0.667327243136242, 0.678086156025529, 0.801198621978983, 0.0462627217639238, 0.772325195139274, 0.349066677736118, 0.31066617066972, 0.233309702947736, 0.270472117001191, 0.073342481860891, 0.225298250326887, 0.131349055794999, 0.176035412587225, 0.797014065552503, 0.466672175331041, 0.926849136129022, 0.873722706455737, 0.713842451572418, 0.202794445678592, 0.773025104543194, 0.725422066403553, 0.000250984914600849, 0.497532946523279, 0.91841639415361, 0.781324994051829, 0.996201495174319, 0.917732512345538, 0.956241200910881, 0.704623289173469, 0.495518396841362, 0.194679045816883, 0.433963091112673, 0.584466230124235, 0.891725917113945, 0.553870343137532, 0.290024423506111, 0.133904293645173, 0.987062872620299, 0.771331703756005, 0.281186288222671, 0.469412123551592, 0.37745515210554, 0.579306234838441, 0.463759727310389, 0.577090732054785, 0.669793729903176, 0.119877251796424, 0.578463905258104, 0.405042729107663, 0.68472212436609, 0.196178160840645, 0.0208894240204245, 0.0298404225613922, 0.286433608736843, 0.030549296643585, 0.724300801521167, 0.839060740545392, 0.547525946749374, 0.607292919186875, 0.389679775107652, 0.23457857593894, 0.0595082642976195, 0.914336876710877, 0.986030810745433, 0.0294006057083607, 0.0249471114948392, 0.531020731199533, 0.479451004415751, 0.114753992529586, 0.725733837345615, 0.536387416534126, 0.814899614080787, 0.910826974315569, 0.603110429132357, 0.483935551252216, 0.778971409425139, 0.358185183489695, 0.711007559671998, 0.346327405422926, 0.711738388752565, 0.564077669521794, 0.660930236103013, 0.407699485076591, 0.0561369168572128, 0.514426220208406, 0.303785680560395, 0.125754167558625, 0.611935834167525, 0.154882293194532, 0.249753871699795, 0.628981029149145, 0.99573928466998, 0.307345874607563, 0.40735580236651, 0.219076152192429, 0.462892433861271, 0.618029822828248, 0.618737344862893, 0.750439253170043, 0.501804614672437, 0.98056960455142, 0.57981473673135, 0.927951003424823, 0.0850992475170642, 0.840990252327174, 0.21337171155028, 0.61642546323128, 0.972262584371492, 0.633642838569358, 0.637900798581541, 0.965201870538294, 0.878753685392439, 0.149407760472968, 0.127427314175293, 0.349494218593463, 0.28139335452579, 0.628005259204656, 0.929732478689402, 0.0683195886667818, 0.252650074427947, 0.694036753149703, 0.704481346532702, 0.918859683908522, 0.672848339891061, 0.420267516979948, 0.766123787499964, 0.866292051039636, 0.497951505472884, 0.208029339555651, 0.976736651035026, 0.0110206508543342, 0.100271746283397, 0.77556214411743, 0.686140236211941, 0.840030050836504, 0.786804303061217, 0.707617257256061, 0.201701588230208, 0.182077475357801, 0.903947841608897, 0.125524511095136, 0.423141689971089, 0.329526523128152, 0.210613703122362, 0.0880531428847462, 0.695514304097742, 0.425684063928202, 0.881715213414282, 0.792732657631859, 0.947299625724554, 0.347086637513712, 0.783139165258035, 0.993491326458752, 0.690737766679376, 0.181901971576735, 0.552987230708823, 0.836811006534845, 0.803252035984769, 0.350695627508685, 0.271387201035395, 0.874111080076545, 0.433307795785367, 0.703320264816284, 0.167030397104099, 0.0392148033715785, 0.964251006487757, 0.0611110748723149, 0.0474057688843459, 0.391849837033078, 0.869372917572036, 0.259487461764365, 0.64685930730775, 0.649622037773952, 0.541786457179114, 0.256748331943527, 0.727916966658086, 0.191629522480071, 0.0787249349523336, 0.34697375074029, 0.563023178838193, 0.125973966438323, 0.797587511362508, 0.909737261477858, 0.250613436568528, 0.706417000154033, 0.501292644999921, 0.948189787101001, 0.985978804994375, 0.426149652805179, 0.515223429305479, 0.47973095625639, 0.963948219781741, 0.767402068478987, 0.311732807895169, 0.591593261808157, 0.619594778399915, 0.105559741845354, 0.331413035746664, 0.734585934784263, 0.0698779881931841, 0.379834534134716, 0.0857429704628885, 0.527639023493975, 0.0208538922015578, 0.224145606160164, 0.239416644209996, 0.682031600503251, 0.00643797940574586, 0.774785183602944, 0.0155732946004719, 0.364181326003745, 0.562782519729808, 0.828557884087786, 0.844410033430904, 0.31228196551092, 0.875076180323958, 0.12835383368656, 0.0599002395756543, 0.421650551026687, 0.664095174754038, 0.785221048630774, 0.769258344080299, 0.922137765446678, 0.590396790299565, 0.237401616061106, 0.728020414011553, 0.592834841925651, 0.854716185480356, 0.666736536193639, 0.15158639010042, 0.0959188791457564, 0.261840453604236, 0.290930520743132, 0.614447548985481, 0.908450851915404, 0.829226288245991, 0.214247517287731, 0.0949732507579029, 0.0145681542344391, 0.0414932733401656, 0.677794784074649, 0.330549358623102, 0.397925784578547, 0.125970720779151, 0.938319414155558, 0.229372118134052, 0.233916701516137, 0.780781991314143, 0.945171435596421, 0.678552735364065, 0.699025796260685, 0.822209337027743, 0.266395327402279, 0.483310784213245, 0.652605385053903, 0.450553887523711, 0.783875183667988, 0.702912505250424, 0.969381730305031, 0.0639264043420553, 0.294088399736211, 0.633087832480669, 0.750613697338849, 0.920038282871246, 0.746032489230856, 0.876996445469558, 0.205956521909684, 0.296554909553379, 0.743854388128966, 0.760806155391037, 0.193779651075602, 0.0480013571213931, 0.0303783707786351, 0.483711007982492, 0.293981208233163, 0.85975272487849, 0.105503575643525, 0.647607798222452, 0.66084094485268, 0.345782415242866, 0.204410893144086, 0.816766494652256, 0.274769733892754, 0.971956607187167, 0.592782198218629, 0.352902981685475, 0.00261648558080196, 0.274698882596567, 0.931024051271379, 0.67943854094483, 0.92263628449291, 0.556158076971769, 0.363713648868725, 0.462250742129982, 0.25597159890458, 0.491215675370768, 0.464514470659196, 0.612759896088392, 0.593527549412102, 0.920706483768299, 0.852793688187376, 0.739342334680259, 0.972731938585639, 0.107305459911004, 0.642250437755138, 0.2203714849893, 0.782261348562315, 0.456184727372602, 0.73975277855061, 0.0362384105101228, 0.392460661241785, 0.226188904372975, 0.126818681135774, 0.172078788280487, 0.0011905871797353, 0.692713270196691, 0.817144309636205, 0.959356692619622, 0.540215470362455, 0.87363129016012, 0.488779205828905, 0.0849958353210241, 0.513371645472944, 0.893392261350527, 0.380524934735149, 0.493107721442357, 0.0911518693901598, 0.107525969622657, 0.317789114546031, 0.299728212412447, 0.378167458577082, 0.382183814188465, 0.411790937883779, 0.493539505405352, 0.0780837689526379, 0.153010735288262, 0.148956429446116, 0.0710461626295, 0.497263734228909, 0.425233855145052, 0.177980433683842, 0.248805993935093, 0.317655288381502, 0.457566024037078, 0.319684850983322, 0.432064145570621, 0.433725903974846, 0.592969113960862, 0.978974359109998, 0.784651955356821, 0.626324718119577, 0.996159727685153, 0.96832575998269, 0.658307396108285, 0.480665899813175, 0.504555382067338, 0.191816122736782, 0.69009796041064, 0.597295890795067, 0.0135236186906695, 0.587417438160628, 0.579724055249244, 0.908500651363283, 0.493643679888919, 0.151720667257905, 0.342119897017255, 0.073768065078184, 0.970582418609411, 0.234308374812827, 0.333553266013041, 0.448073349194601, 0.837035174015909, 0.48441000119783, 0.380536791868508, 0.802949814824387, 0.655068137682974, 0.0699601031374186, 0.274038116447628, 0.171953214099631, 0.26745026698336, 0.948718541534618, 0.516780630219728, 0.507348821731284, 0.440953947836533, 0.285967678530142, 0.498949200147763, 0.255292442394421, 0.267607634654269, 0.220636613667011, 0.630345017882064, 0.944121848093346, 0.832299566827714, 0.081855736207217, 0.175691784359515, 0.0573780431877822, 0.567132550058886, 0.875628162641078, 0.176033878931776, 0.445413268869743, 0.248295004013926, 0.0548252346925437, 0.495049024233595, 0.133154788287356, 0.883212363580242, 0.030415453016758, 0.26593574648723, 0.963179147569463, 0.812917772913352, 0.805255657993257, 0.882563073420897, 0.495936438441277, 0.0629192644264549, 0.437602991703898, 0.115770126227289, 0.814004107145593, 0.544575568055734, 0.220733462367207, 0.377239837544039, 0.821336790453643, 0.114755036542192, 0.498979869997129, 0.602100278716534, 0.644885292043909, 0.470144180580974, 0.127464689314365, 0.700432159472257, 0.525933096883819, 0.566998125985265, 0.371458732755855, 0.758648809976876, 0.854866087669507, 0.681870767846704, 0.903857962694019, 0.787022436968982, 0.365754841128364, 0.580766278551891, 0.872923610266298, 0.377662983490154, 0.312065861420706, 0.635844451142475, 0.864610915305093, 0.231578717241064, 0.659971904009581, 0.536052697570994, 0.538513367297128, 0.72389054694213, 0.0123669491149485, 0.567255374277011, 0.149086055811495, 0.116909028496593, 0.256700576283038, 0.32775421324186, 0.927484668092802, 0.876048493431881, 0.380962274968624, 0.513893277850002, 0.667537309229374, 0.582592588849366, 0.586811058223248, 0.264097207924351, 0.71094482159242, 0.253959578927606, 0.730408302275464, 0.26138222636655, 0.26375852455385, 0.545347794191912, 0.591474881861359, 0.941070346860215, 0.968028383795172, 0.900958501268178, 0.443880125181749, 0.410049551632255, 0.706304290797561, 0.520523368846625, 0.732102599693462, 0.0356005784124136, 0.581528341164812, 0.317270398139954, 0.802927945740521, 0.0902990086469799, 0.65061583626084, 0.139798269141465, 0.00540950195863843, 0.0420476878061891, 0.998304752632976, 0.655676488531753, 0.467146837385371, 0.589249668875709, 0.0615844805724919, 0.466565191512927, 0.730742913903669, 0.773404429201037, 0.47816896578297, 0.385106170084327, 0.0675250368658453, 0.0369439488276839, 0.616421923041344, 0.967991313897073, 0.665331850759685, 0.0660395426675677, 0.719516516197473, 0.51928139408119, 0.0136446424294263, 0.992840198799968, 0.490160649409518, 0.447134271264076, 0.731534676859155, 0.880416043102741, 0.822674663504586, 0.765696871560067, 0.0300758895464242, 0.765857407590374, 0.822555248625576, 0.31150103919208, 0.414386169286445, 0.333247975911945, 0.0484359168913215, 0.712400741875172, 0.640355089912191, 0.432713193586096, 0.778612199705094, 0.240928122540936, 0.518919319612905, 0.664381850278005, 0.324887358816341, 0.431436370592564, 0.952359968796372, 0.273197899339721, 0.9753819166217, 0.100576792145148, 0.438784726429731, 0.488198173465207, 0.425137630896643, 0.930220771348104, 0.398630132200196, 0.647306388942525, 0.0560423820279539, 0.431326035410166, 0.682732199085876, 0.0570194036699831, 0.430339703103527, 0.534638341050595, 0.296387939713895, 0.0294658928178251, 0.311050365678966, 0.0220108043868095, 0.647186506073922, 0.61523867608048, 0.902677448000759, 0.149653801228851, 0.049387977225706, 0.0286969926673919, 0.792663235915825, 0.156522075179964, 0.346922914264724, 0.641889060847461, 0.794229435734451, 0.113093732623383, 0.191193169448525, 0.367569910595194, 0.344270161353052, 0.16332206199877, 0.444721235660836, 0.811371711082757, 0.337223218288273, 0.815041722729802, 0.499245872488245, 0.0633981011342257, 0.693950133398175, 0.791797860292718, 0.807519927620888, 0.0349234861787409, 0.546711799921468, 0.311885839095339, 0.638949579559267, 0.540017252555117, 0.450933414744213, 0.198016109643504, 0.301151155494153, 0.725895224837586, 0.464056733995676, 0.564365660306066, 0.533237151801586, 0.378347933059558, 0.0769829703494906, 0.352863384410739, 0.640958827687427, 0.13786181807518, 0.590888299280778, 0.717560434248298, 0.868939579464495, 0.0218992361333221, 0.759659278672189, 0.375753603875637, 0.805280677974224, 0.719421321759, 0.60262307850644, 0.985746888443828, 0.649327573599294, 0.406610208097845, 0.847021056571975, 0.466033183271065, 0.492641815915704, 0.681470410199836, 0.112184192053974, 0.728494939627126, 0.761197955813259, 0.19780385447666, 0.156875653658062, 0.259477833285928, 0.956473858095706, 0.480331061407924, 0.0630517846439034, 0.122619517380372, 0.723063841694966, 0.873786664102226, 0.963705749250948, 0.354584710672498, 0.80872478755191, 0.821595007088035, 0.241914940997958, 0.550975744379684, 0.235282499575987, 0.440830728737637, 0.654209074797109, 0.675865029683337, 0.842861314071342, 0.594433441292495, 0.192565185949206, 0.986546129686758, 0.725121404277161, 0.361988575663418, 0.403515809448436, 0.478787173284218, 0.146808101562783, 0.99422034714371, 0.939436520449817, 0.25282798986882, 0.556972639635205, 0.417464949190617, 0.974866850301623, 0.433143001748249, 0.38641393953003, 0.137641547713429, 0.602190774399787, 0.874907130608335, 0.670062082819641, 0.715189126320183, 0.35900089959614, 0.4473059582524, 0.675532911671326, 0.327581462683156, 0.629905396373942, 0.675955653423443, 0.482290443964303, 0.79972408246249, 0.95165110914968, 0.516451804665849, 0.43233325262554, 0.887016190681607, 0.965704360976815, 0.278498373925686, 0.170195486629382, 0.0752366622909904, 0.373807456111535, 0.895414605969563, 0.32270574499853, 0.460041711106896, 0.157765539130196, 0.0692700115032494, 0.39208030840382, 0.0615924748126417, 0.696437316481024, 0.911976000759751, 0.397681509144604, 0.529840913135558, 0.232790913898498, 0.682165423408151, 0.088888131082058, 0.0596892633475363, 0.386474082944915, 0.0956045102793723, 0.977478964719921, 0.924487638054416, 0.417934696190059, 0.294459854951128, 0.524983893614262, 0.936562744667754, 0.62559163197875, 0.966604890068993, 0.358429925283417, 0.346304569859058, 0.857705094618723, 0.778367427410558, 0.612705336883664, 0.48939023562707, 0.222902682144195, 0.602343816077337, 0.322972465772182, 0.0181337941903621, 0.694680789019912, 0.224504990968853, 0.903675383655354, 0.515649042790756, 0.103550459025428, 0.517777189379558, 0.553413408109918, 0.663558864966035, 0.857112180208787, 0.83712908742018, 0.944573547458276, 0.806032251333818, 0.672103882068768, 0.577460383763537, 0.690344421891496, 0.66367589822039, 0.273694592062384, 0.459667304297909, 0.0140643750783056, 0.992049682186916, 0.780239338520914, 0.755377281690016, 0.46210939460434, 0.554693377809599, 0.322720916476101, 0.981678078882396, 0.431060984730721, 0.617897980613634, 0.25859724637121, 0.114596440456808, 0.975091321393847, 0.441285689128563, 0.830607168842107, 0.146333591314033, 0.742433121660724, 0.235829015029594, 0.89435369707644, 0.428355748998001, 0.142838031752035, 0.744855475146323, 0.190786578692496, 0.582708646077663, 0.129001806955785, 0.616629738127813, 0.0259906351566315, 0.939714380307123, 0.0304916379973292, 0.0143524701707065, 0.947575885569677, 0.181943941162899, 0.0491979252547026, 0.446298765717074, 0.81704317452386, 0.991720995632932, 0.445392054738477, 0.844360508024693, 0.777429847745225, 0.926040043355897, 0.273347232025117, 0.012316630454734, 0.7457763585262, 0.585420752409846, 0.486011347034946, 0.0360604431480169, 0.649935488821939, 0.924721217248589, 0.300640014698729, 0.703191800042987, 0.555942408274859, 0.587785464245826, 0.646012328565121, 0.9805155582726, 0.0966857364401221, 0.0300111998803914, 0.672418980160728, 0.0102989901788533, 0.719964930554852, 0.202015593647957, 0.607230279361829, 0.702213309239596, 0.0752659267745912, 0.837762464070693, 0.731334941461682, 0.761705259792507, 0.327977841952816, 0.918029482942075, 0.687488268595189, 0.942026868928224, 0.882982631446794, 0.387390905991197, 0.851302118506283, 0.580577423097566, 0.41453733551316, 0.82214065734297, 0.0831012239214033, 0.113474159734324, 0.606126158963889, 0.0398917647544295, 0.458086827537045, 0.83261967333965, 0.821552301757038, 0.747359573375434, 0.801502056419849, 0.995249710278586, 0.69111097836867, 0.549706236459315, 0.0798317920416594, 0.282085625454783, 0.651608423795551, 0.702794924844056, 0.20111318747513, 0.609428152674809, 0.0576419532299042, 0.0667067659087479, 0.0558077187743038, 0.397291867760941, 0.642388238571584, 0.722493647132069, 0.0498933950439095, 0.603744288906455, 0.117423413088545, 0.299385889200494, 0.830656599719077, 0.195220710011199, 0.421630071708933, 0.281036452157423, 0.812799915904179, 0.796218886040151, 0.670095805078745, 0.977968697901815, 0.907327926484868, 0.951028402196243, 0.900983929401264, 0.968951254617423, 0.297398682218045, 0.429706631228328, 0.489473577355966, 0.793388617224991, 0.384024490369484, 0.999046531505883, 0.595917958999053, 0.883023581467569, 0.646617607446387, 0.734924722928554, 0.342680286616087, 0.681803205283359, 0.583000744692981, 0.744304427877069, 0.821730113355443, 0.156843676697463, 0.122910924488679, 0.500524047296494, 0.986990581033751, 0.0816238902043551, 0.486428667092696, 0.703278628643602, 0.284937924705446, 0.461537003982812, 0.565920985070989, 0.651398652233183, 0.259219079278409, 0.0815291756298393, 0.297071366338059, 0.44653339125216, 0.57284515700303, 0.243276887806132, 0.262899844907224, 0.840631179744378, 0.403060441603884, 0.0519369223620743, 0.697961266385391, 0.544435408897698, 0.857315367087722, 0.545826596207917, 0.201033737743273, 0.529765206156299, 0.682584576308727, 0.799741948954761, 0.0311837280169129, 0.132999405264854, 0.176701874705032, 0.478745598345995, 0.0753256897442043, 0.34063389711082, 0.613732848782092, 0.688057579565793, 0.0811311728321016, 0.137850729981437, 0.943386592902243, 0.0485985763370991, 0.990808357018977, 0.326636982616037, 0.10908980644308, 0.627875644247979, 0.452355381567031, 0.589865481713787, 0.523429250344634, 0.105533802183345, 0.744886012980714, 0.25263167754747, 0.189690316095948, 0.702858899487182, 0.659670326858759, 0.785275149159133, 0.988217320060357, 0.362718340707943, 0.361412284895778, 0.337886887136847, 0.589417080860585, 0.707644048845395, 0.918094431748614, 0.495908894110471, 0.745811824686825), actual = c(1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"