Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.7571020        0.1133799 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5547456        0.5547456        0.5118995 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.573734054109082, 0.546838962007314, 0.0679954118095338, 0.822887259768322, 0.286676428513601, 0.857261137105525, 0.0782489189878106, 0.803828374948353, 0.710226067807525, 0.666532583301887, 0.747410361189395, 0.695049981353804, 0.703115678392351, 0.710564914392307, 0.407591908704489, 0.920575607335195, 0.568226398667321, 0.687493125209585, 0.385663452092558, 0.622728512389585, 0.172079534269869, 0.219741690671071, 0.824171141721308, 0.772397487657145, 0.228220423450693, 0.688181242439896, 0.850855776807293, 0.994307292625308, 0.32753030047752, 0.125749542145059, 0.547281438019127, 0.192260354757309, 0.202693476341665, 0.87762212427333, 0.176294029224664, 0.866533382795751, 0.014820106793195, 0.32802245952189, 0.73434257786721, 0.875983021687716, 0.67656339192763, 0.675153117859736, 0.975987015292048, 0.206448011798784, 0.241014046594501, 0.153398924507201, 0.0154428521636873, 0.181681773159653, 0.604782480979338, 0.937870705733076, 0.656059411587194, 0.546841465169564, 0.60518142557703, 0.884779290994629, 0.690801969263703, 0.0836242395453155, 0.110036680009216, 0.709567475598305, 0.484399798326194, 0.849969224305823, 0.981459161499515, 0.533224648563191, 0.291699627181515, 0.362103212159127, 0.689389682840556, 0.60362021625042, 0.74423092417419, 0.14980333019048, 0.864227706566453, 0.11478428193368, 0.285665269009769, 0.802257226780057, 0.318483362207189, 0.574942447710782, 0.669493827270344, 0.107198046986014, 0.311339573701844, 0.933629463426769, 0.309139367425814, 0.00590781844221056, 0.891792266396806, 0.842039259616286, 0.29304024297744, 0.851554198889062, 0.793318211333826, 0.909184881253168, 0.745929024880752, 0.281484568025917, 0.540062984451652, 0.858396545983851, 0.114740198710933, 0.128839988727123, 0.668193072779104, 0.270894133020192, 0.666465111076832, 0.856346861692145, 0.786741968942806, 0.453137093922123, 0.531078499509022, 0.103364879032597, 0.507287666667253, 0.70696306717582, 0.835704050026834, 0.83053092751652, 0.815861774841323, 0.465717830462381, 0.0483631989918649, 0.982873782282695, 0.236367563949898, 0.275076552294195, 0.632073588902131, 0.0301804835908115, 0.844289110740647, 0.969289044383913, 0.741728516062722, 0.90827525826171, 0.408233545487747, 0.785728230373934, 0.37480899412185, 0.457345070317388, 0.0649473294615746, 0.58207274088636, 0.597812236053869, 0.460377833340317, 0.465888711391017, 0.85419784206897, 0.305779130430892, 0.0786770542617887, 0.0822823287453502, 0.842444165144116, 0.229399250121787, 0.0673778888303787, 0.951997705036774, 0.26983853103593, 0.3073239417281, 0.61699055088684, 0.439751767553389, 0.565080180764198, 0.177160752704367, 0.546849037287757, 0.82801450509578, 0.708303224062547, 0.0892240584362298, 0.899435297353193, 0.58941082819365, 0.309328146511689, 0.642683226149529, 0.246510475873947, 0.0546555230394006, 0.135119949001819, 0.196349409874529, 0.647000572877005, 0.102420203620568, 0.759900567121804, 0.0889370711520314, 0.311974304262549, 0.44446656270884, 0.52736240089871, 0.484165939735249, 0.891421631211415, 0.33089815126732, 0.0697186749894172, 0.770163919078186, 0.462883832165971, 0.834548395592719, 0.37620426248759, 0.0327878340613097, 0.647933452390134, 0.737886781804264, 0.546858122106642, 0.9716006454546, 0.529946056427434, 0.863847483647987, 0.102055630180985, 0.89330889377743, 0.931893835077062, 0.186033993726596, 0.144293361809105, 0.996040508616716, 0.771642369916663, 0.124208625406027, 0.171724527841434, 0.909703599987552, 0.0550311550032347, 0.907175640109926, 0.823311690473929, 0.21958045498468, 0.859790266957134, 0.967448891839013, 0.801511765224859, 0.572096907999367, 0.488641391973943, 0.0438664180692285, 0.10198513395153, 0.155147747602314, 0.21151848626323, 0.446158500621095, 0.0946862325072289, 0.714258806314319, 0.254911112366244, 0.0659474402200431, 0.762354549020529, 0.232121490174904, 0.916222571395338, 0.327973776031286, 0.68732169480063, 0.48836334538646, 0.815644989255816, 0.437852960079908, 0.802467555273324, 0.303530141245574, 0.822723905788735, 0.218153955182061, 0.961437499383464, 0.199661823455244, 0.0781408345792443, 0.702871136832982, 0.0899512120522559, 0.893529392080382, 0.705341859487817, 0.682044322602451, 0.628242897568271, 0.615081781987101, 0.177292925538495, 0.0672721799928695, 0.726809087907895, 0.995853708125651, 0.716057769255713, 0.217843282967806, 0.841143800644204, 0.238972414750606, 0.352319131605327, 0.473399539943784, 0.403307085623965, 0.904954309342429, 0.0143431406468153, 0.350505019770935, 0.571420820197091, 0.658232303801924, 0.12096939701587, 0.281040039844811, 0.836833892390132, 0.146968148415908, 0.148792156251147, 0.652328780619428, 0.610050922492519, 0.778141993563622, 0.360155376140028, 0.540092953946441, 0.514426123118028, 0.84544545901008, 0.64440081268549, 0.285097041167319, 0.68883721739985, 0.888119490584359, 0.370031846687198, 0.771170142805204, 0.156563095282763, 0.286362240323797, 0.607273577712476, 0.993401013314724, 0.258608995936811, 0.0280736875720322, 0.78312323265709, 0.80971982027404, 0.993118691723794, 0.964365085121244, 0.65772822056897, 0.225406877929345, 0.431278721895069, 0.451176440576091, 0.720632455777377, 0.278172636404634, 0.985560538712889, 0.450059234397486, 0.0820552567020059, 0.353918889304623, 0.26044455380179, 0.48245142865926, 0.372476503020152, 0.654900696594268, 0.274137862958014, 0.999398683197796, 0.0416599637828767, 0.0820742323994637, 0.933697840431705, 0.991877654800192, 0.312023667851463, 0.973382009891793, 0.353586874902248, 0.704291533911601, 0.195074068149552, 0.213716347236186, 0.237444131635129, 0.551636741030961, 0.67030843347311, 0.492888041539118, 0.378158911829814, 0.515501317568123, 0.621784257469699, 0.773252255283296, 0.549163959687576, 0.514358521904796, 0.823609416605905, 0.146098830504343, 0.72005166602321, 0.804294996662065, 0.470253870589659, 0.19570410810411, 0.196273035835475, 0.326633722055703, 0.518966875737533, 0.214032572461292, 0.430607025511563, 0.502082380000502, 0.270579433534294, 0.98221652302891, 0.856118990806863, 0.561471901135519, 0.728055071318522, 0.197117363335565, 0.942923066439107, 0.310644207056612, 0.214836707338691, 0.277262173360214, 0.837546652648598, 0.442358747823164, 0.851235433248803, 0.821538117481396, 0.941733165178448, 0.657076182309538, 0.921874094055966, 0.765514096012339, 0.338628143072128, 0.522522746352479, 0.28099626256153, 0.0255577964708209, 0.218044882873073, 0.384732965612784, 0.964864889858291, 0.771609820891172, 0.176508557749912, 0.511325482977554, 0.995893647428602, 0.958994500571862, 0.395594061119482, 0.683883026242256, 0.375329659087583, 0.762225288664922, 0.887441661208868, 0.847165467683226, 0.768316195812076, 0.997612881241366, 0.301633154973388, 0.178670218447223, 0.289253246504813, 0.901011666748673, 0.290832480415702, 0.279538709437475, 0.0561901486944407, 0.0246646699961275, 0.584450730355456, 0.288884837180376, 0.191463289316744, 0.441934107104316, 0.96871778415516, 0.998550510266796, 0.232051897794008, 0.581908549414948, 0.28542599780485, 0.961366944946349, 0.00405818433500826, 0.322694660164416, 0.559006663737819, 0.967779272701591, 0.185790832387283, 0.468557629967108, 0.506176310358569, 0.982388878474012, 0.943097100127488, 0.0598974931053817, 0.920230276184157, 0.31835554074496, 0.891456910176203, 0.728898416506127, 0.504722017794847, 0.0807145070284605, 0.0225942668039352, 0.352273343829438, 0.799483892740682, 0.13917937874794, 0.0854098750278354, 0.400679919403046, 0.00451159826479852, 0.473072326276451, 0.197440865682438, 0.98082689428702, 0.192334704799578, 0.973015225259587, 0.421192022040486, 0.0214218408800662, 0.451287446776405, 0.251246440922841, 0.422366316895932, 0.0945058693177998, 0.196364750852808, 0.476501576369628, 0.184815694345161, 0.190975338919088, 0.417341274907812, 0.923107186565176, 0.795450859703124, 0.803932214388624, 0.266966399038211, 0.888158654561266, 0.658311595441774, 0.190991504350677, 0.416156442370266, 0.503494131844491, 0.722407296532765, 0.637325819581747, 0.337516671046615, 0.890052882721648, 0.725812180899084, 0.495711505645886, 0.152767952764407, 0.0343816268723458, 0.148338784230873, 0.817402448039502, 0.0687922623474151, 0.118591176113114, 0.283024898963049, 0.442473279777914, 0.857375014107674, 0.838663501897827, 0.725585896288976, 0.269518679473549, 0.217094548046589, 0.639840634772554, 0.534894752781838, 0.987993058981374, 0.633365378947929, 0.314557472942397, 0.696247943677008, 0.0913990195840597, 0.56720426841639, 0.198176475707442, 0.441154835978523, 0.405204919865355, 0.432613840792328, 0.914827014319599, 0.0340815591625869, 0.697858939180151, 0.625854445621371, 0.154694721801206, 0.417521730996668, 0.19935419363901, 0.106348887318745, 0.585686980979517, 0.507719064829871, 0.996603962499648, 0.904852845473215, 0.96549378009513, 0.690719428937882, 0.449323042063043, 0.267093147151172, 0.692909135017544, 0.107712231809273, 0.844020517077297, 0.685054572997615, 0.865125159732997, 0.951955473748967, 0.0168043761514127, 0.703438937198371, 0.796648861374706, 0.0994963375851512, 0.735468559665605, 0.358823655871674, 0.708067235304043, 0.445900853257626, 0.73846283601597, 0.500031359959394, 0.13375225965865, 0.0665743546560407, 0.392287977971137, 0.253446166869253, 0.887877721339464, 0.791968320962042, 0.62183212232776, 0.296073538484052, 0.944961621891707, 0.294794106390327, 0.781193725764751, 0.0812082197517157, 0.318451306084171, 0.667867588344961, 0.862233457621187, 0.726711951196194, 0.924705982673913, 0.0356458185706288, 0.674674261361361, 0.318625922547653, 0.546589396661147, 0.588939778273925, 0.945525377290323, 0.661230540834367, 0.360717218834907, 0.0617219447158277, 0.80342804430984, 0.536612931871787, 0.359910330967978, 0.942882624221966, 0.865803104825318, 0.204683807678521, 0.0724817202426493, 0.0860688975080848, 0.561025937320665, 0.476869135396555, 0.177510536974296, 0.817561497911811, 0.486579449148849, 0.819353784900159, 0.551066235639155, 0.790599209954962, 0.213080117944628, 0.256219509290531, 0.199687422486022, 0.16040376201272, 0.396155394148082, 0.691704038297758, 0.223526891320944, 0.977653047302738, 0.523388540372252, 0.805945615749806, 0.0335684982128441, 0.152430956950411, 0.612534434301779, 0.402029276825488, 0.497573006199673, 0.591620217775926, 0.589885503053665, 0.165627273730934, 0.367499950574711, 0.689931705128402, 0.906158853089437, 0.972153129288927, 0.738751532509923, 0.47128137643449, 0.541331517277285, 0.870522070676088, 0.423423312837258, 0.627999060088769, 0.829598056385294, 0.0182440881617367, 0.475434743799269, 0.95023038238287, 0.106214771512896, 0.338727943366393, 0.989368249895051, 0.165877272840589, 0.64174691401422, 0.668117907363921, 0.897644789190963, 0.616816682275385, 0.296427299734205, 0.616150570567697, 0.988743678666651, 0.926423303084448, 0.658210358116776, 0.402294662082568, 0.308265361003578, 0.652762402314693, 0.748587385518476, 0.461601031245664, 0.00872101681306958, 0.319926536176354, 0.0770298454444855, 0.151259836275131, 0.984870171174407, 0.35679767257534, 0.122449018759653, 0.45319961477071, 0.838200256461278, 0.346251191105694, 0.210712707834318, 0.745720260776579, 0.921798540279269, 0.953780883690342, 0.154309931676835, 0.0267836984712631, 0.513183604925871, 0.799627600237727, 0.743053498445079, 0.405735326465219, 0.94214819278568, 0.860847175121307, 0.552665946539491, 0.921634596306831, 0.744118103059009, 0.510548112681136, 0.74657694878988, 0.470135923707858, 0.0988193748053163, 0.0951627513859421, 0.570594884920865, 0.502061216393486, 0.249920283211395, 0.571844537742436, 0.00944004766643047, 0.895664937561378, 0.727531377924606, 0.846853787545115, 0.304421374807134, 0.157205484108999, 0.576780213974416, 0.931615963578224, 0.0860870301257819, 0.640296465018764, 0.852573645301163, 0.528469062643126, 0.239419164368883, 0.492209903895855, 0.821061206050217, 0.945319376885891, 0.665293633705005, 0.33048653928563, 0.505185086745769, 0.497970586642623, 0.0727345631457865, 0.443504002643749, 0.231423046905547, 0.662343450821936, 0.972118028206751, 0.551541031803936, 0.258347058435902, 0.247097690356895, 0.400545275304466, 0.504005655646324, 0.530588517664, 0.651053942041472, 0.862588745076209, 0.951299156993628, 0.0729801401030272, 0.570535408565775, 0.224179691169411, 0.434831937542185, 0.485498243942857, 0.13911583670415, 0.149525819346309, 0.0344823945779353, 0.108766748802736, 0.298146209446713, 0.671048929682001, 0.0344213400967419, 0.874924835981801, 0.53969888901338, 0.255001634126529, 0.264330990146846, 0.320939609548077, 0.17206471785903, 0.133956603938714, 0.582432622555643, 0.485734989866614, 0.871615623822436, 0.00750760454684496, 0.0107523233164102, 0.170246405992657, 0.64096268452704, 0.964935369323939, 0.364469807362184, 0.592443897388875, 0.399332833476365, 0.394270349061117, 0.0382105601020157, 0.198392061050981, 0.24380056001246, 0.28203082154505, 0.69107844796963, 0.609873403329402, 0.0809285012073815, 0.455449776956812, 0.801367022795603, 0.912125411443412, 0.637443190906197, 0.168936438160017, 0.759897269774228, 0.884318261407316, 0.64076529094018, 0.493363360175863, 0.425231749890372, 0.34613527264446, 0.384529170813039, 0.624249427113682, 0.198510349728167, 0.64323217747733, 0.543248290661722, 0.622107295319438, 0.984158645616844, 0.0185717653948814, 0.407297682249919, 0.75473789870739, 0.270799501566216, 0.209988252259791, 0.987889570649713, 0.453667683526874, 0.579363488592207, 0.0213625046890229, 0.0399883582722396, 0.631767350248992, 0.436075557488948, 0.537353368243203, 0.759101647417992, 0.921601310372353, 0.953234502114356, 0.195271326461807, 0.111366916447878, 0.0699019089806825, 0.743334389291704, 0.250834135804325, 0.906647981377319, 0.0101809797342867, 0.549114492256194, 0.734562622848898, 0.313883294351399, 0.61542468238622, 0.933442519977689, 0.905081705423072, 0.312864942010492, 0.435844351071864, 0.502025209832937, 0.293555441079661, 0.0168509271461517, 0.131358248647302, 0.727878934470937, 0.629268234595656, 0.641812266781926, 0.604075297247618, 0.644912832649425, 0.237446045735851, 0.0237947660498321, 0.712835759157315, 0.356605054344982, 0.381743845995516, 0.198717691237107, 0.90193081763573, 0.298556519672275, 0.0439060821663588, 0.999781294027343, 0.150358648737893, 0.10385138890706, 0.0861463169567287, 0.745541343232617, 0.294120243517682, 0.522423032904044, 0.214759243186563, 0.322750615887344, 0.641897729830816, 0.48199573927559, 0.212019188329577, 0.0567766530439258, 0.729785766918212, 0.745034547755495, 0.518590503605083, 0.89250727253966, 0.873146899510175, 0.643299594987184, 0.59830603399314, 0.880952737759799, 0.112785403616726, 0.092715535312891, 0.460019933525473, 0.788318080361933, 0.862646972993389, 0.810117612825707, 0.317664911039174, 0.368258700007573, 0.296562893083319, 0.94753379910253, 0.149004505714402, 0.727025249274448, 0.637594711501151, 0.0421618677210063, 0.448313619475812, 0.0965421323198825, 0.164278307929635, 0.316605142550543, 0.713201859267429, 0.628047143574804, 0.327326751546934, 0.0396065679378808, 0.204910419415683, 0.371929036686197, 0.11284447950311, 0.0977557408623397, 0.783807722851634, 0.11494296323508, 0.988760272273794, 0.778668378479779, 0.677039967384189, 0.978504162514582, 0.340807921485975, 0.46455966681242, 0.656299531459808, 0.402440381469205, 0.629493124084547, 0.97520600259304, 0.123920451616868, 0.86047184234485, 0.421436560340226, 0.241176193580031, 0.953830727608874, 0.326736881863326, 0.127931343624368, 0.382861864287406, 0.104594193398952, 0.877098427154124, 0.4803085278254, 0.164033072767779, 0.789434387581423, 0.254636527271941, 0.314630483277142, 0.500980144366622, 0.931985270231962, 0.499979971675202, 0.939776226645336, 0.282126255333424, 0.683947963640094, 0.250880345003679, 0.288082074839622, 0.504981085192412, 0.340612461557612, 0.680321246152744, 0.602226895978674, 0.907420953502879, 0.533182747196406, 0.158454998861998, 0.154119750484824, 0.965687453281134, 0.728487676475197, 0.353959484258667, 0.767323142848909, 0.668438660679385, 0.872247279155999, 0.0273648865986615, 0.0323627756442875, 0.275993501301855, 0.848926316248253, 0.691458718618378, 0.321918487315997, 0.294301439309493, 0.804099639877677, 0.355821652105078, 0.708149346755818, 0.109272954054177, 0.495792233385146, 0.605455198790878, 0.459481393219903, 0.763443120056763, 0.661059873411432, 0.392184884520248, 0.0631185793317854, 0.561346478294581, 0.301022398518398, 0.46616780012846, 0.982106218347326, 0.00846484326757491, 0.963123920606449, 0.996801078086719, 0.231807769974694, 0.594917366746813, 0.220605633687228, 0.04301393404603, 0.61481447191909, 0.186483148951083, 0.126246484927833, 0.778677474940196, 0.1325833783485, 0.675075566163287, 0.300431563984603, 0.361037112073973, 0.222966957837343, 0.260053907055408, 0.672163339331746, 0.479217960499227, 0.348553451476619, 0.965372080216184, 0.810359386960045, 0.222599255852401, 0.289529251633212, 0.210656168637797, 0.736978904809803, 0.469198051374406, 0.0578912438359112, 0.467995798448101, 0.773990902351215, 0.21539840195328, 0.424966769991443, 0.0631119937170297, 0.641431941650808, 0.43813694966957, 0.806090469937772, 0.192205850733444, 0.593406048836187, 0.846454882761464, 0.240800209576264, 0.447980806231499, 0.544397608842701, 0.38539391499944, 0.832572306971997, 0.86991616548039, 0.452230669558048, 0.670308427186683, 0.403798236977309, 0.49253362044692, 0.457316734595224, 0.984180180588737, 0.128729359945282, 0.10032558417879, 0.604970968561247, 0.329345436533913, 0.100015839794651, 0.132709398400038, 0.998485522577539, 0.851590370759368, 0.349467676365748, 0.317360091023147, 0.593247983371839, 0.0834016888402402, 0.636693073669448, 0.186867865733802, 0.966871234122664, 0.0425670479889959, 0.614498548675328, 0.287747641094029, 0.600603153929114, 0.398123312974349, 0.145954767474905, 0.546416835859418, 0.367825432913378, 0.943063595565036, 0.330136892385781, 0.165973246097565, 0.47509980481118, 0.214262600755319, 0.711941661313176, 0.12767382664606, 0.666458030929789, 0.116968742571771, 0.459083227207884, 0.593787897145376, 0.768433840246871, 0.844766042428091, 0.359398829285055, 0.326097738929093, 0.152681287145242, 0.665729655185714, 0.987511598970741, 0.3520786925219, 0.74895351100713, 0.37140781339258, 0.625857323175296, 0.259403347969055, 0.888901326106861, 0.512983270687982, 0.31705762934871, 0.52684959769249, 0.818670882610604, 0.120056610787287, 0.740456230472773, 0.209772927220911, 0.394822714850307, 0.840171626303345, 0.510016758693382, 0.414057268993929, 0.483455941081047, 0.379643089370802, 0.826423988677561, 0.127874413738027, 0.0422613762784749, 0.63673720904626, 0.17018110700883, 0.580910918070003, 0.380259279161692, 0.832170173060149, 0.329270559130237, 0.0590890343300998, 0.711973461322486, 0.802184835309163, 0.964668314205483, 0.718825618736446, 0.777301901718602, 0.901712959399447, 0.189358121249825, 0.298688215902075, 0.667805891018361, 0.357120998203754, 0.0729152380954474, 0.560375745408237, 0.902122944127768, 0.723562679486349), actual = c(0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.0874406085349619, 0.767411150271073, 0.528758869739249, 0.969298487529159, 0.945628154790029, 0.944109271746129, 0.754960541380569, 0.934212918858975, 0.169159951154143, 0.573194256983697, 0.825328350067139, 0.443164314609021, 0.677032248582691, 0.663968209875748, 0.540601900080219, 0.24924526643008, 0.989144874969497, 0.908391109434888, 0.973290400346741, 0.60539279691875, 0.706691501429304, 0.0719714048318565, 0.238787712296471, 0.237149958265945, 0.529041685163975, 0.854504622519016, 0.967750764917582, 0.731883329572156, 0.713666185038164, 0.610396043164656, 0.866450431058183, 0.791958438465372, 0.368235821835697, 0.916419463697821, 0.700179744744673, 0.0413162556942552, 0.482731145806611, 0.80168288666755, 0.154036739375442, 0.0148754788096994, 0.453660408966243, 0.66494298982434, 0.679675192572176, 0.451275702565908, 0.626076931599528, 0.783029899466783, 0.766065107425675, 0.426180174807087, 0.464166342979297, 0.992508918046951, 0.19583057006821, 0.622843354707584, 0.409760983893648, 0.951330240815878, 0.572261442430317, 0.763281293446198, 0.0518453612457961, 0.433250319678336, 0.0960037228651345, 0.103331216378137, 0.256588172866032, 0.572415471542627, 0.47162489569746, 0.230493793729693, 0.406242218799889, 0.353913426864892, 0.112102180719376, 0.14079236285761, 0.220186705002561, 0.305685917846859, 0.715067126089707, 0.617371672764421, 0.689532596385106, 0.461480908328667, 0.626295268069953, 0.893215301679447, 0.354108703322709, 0.849452382652089, 0.280605213250965, 0.657543022884056, 0.352000718470663, 0.430270199896768, 0.727788703050464, 0.22718525188975, 0.366747750435024, 0.312040809541941, 0.907656291034073, 0.54710047529079, 0.130053747911006, 0.549254336161539, 0.0935997450724244, 0.573060925118625, 0.0202793313656002, 0.842133164871484, 0.873188246507198, 0.749120177468285, 0.100826385896653, 0.912755999481305, 0.872838966781273, 0.68142073508352, 0.474238649243489, 0.666278192074969, 0.931061968673021, 0.428647074382752, 0.565069752279669, 0.27197370887734, 0.659093830035999, 0.032979000126943, 0.701213506050408, 0.623968559317291, 0.316140135517344, 0.382763552013785, 0.0454080407507718, 0.356427515624091, 0.820704880636185, 0.405390109401196, 0.900450965389609, 0.170109277358279, 0.212003579596058, 0.311152331996709, 0.128927068784833, 0.355868702987209, 0.121797050815076, 0.478940531145781, 0.511256356257945, 0.443488651653752, 0.0642134675290436, 0.315977177349851, 0.0248096052091569, 0.969988072523847, 0.816195613471791, 0.969418973196298, 0.159705867292359, 0.0472488389350474, 0.799990087747574, 0.878506915410981, 0.705315723549575, 0.566877518314868, 0.0591700468212366, 0.736393021419644, 0.452300236793235, 0.0421579184476286, 0.122695808997378, 0.40058483555913, 0.520571895409375, 0.811962601961568, 0.815286997472867, 0.136005422798917, 0.406424931250513, 0.5231615123339, 0.630967562552541, 0.843603338580579, 0.693893119459972, 0.235776650952175, 0.705804353812709, 0.436273806961253, 0.129515565000474, 0.720294102095068, 0.624317630659789, 0.231984719634056, 0.322536092950031, 0.627929769689217, 0.97321894345805, 0.658102361951023, 0.233479707734659, 0.98543878365308, 0.583330345107242, 0.5457127855625, 0.0455254619009793, 0.360121819190681, 0.637734189629555, 0.524300230201334, 0.458306376589462, 0.586612351238728, 0.439756489358842, 0.485099812969565, 0.74246304994449, 0.690355309285223, 0.65954097523354, 0.0706204771995544, 0.718219221336767, 0.758615651866421, 0.119989601895213, 0.668951661558822, 0.493728856090456, 0.249558598501608, 0.703340256586671, 0.192589788231999, 0.808732756879181, 0.714660171652213, 0.254932298790663, 0.385708520654589, 0.472851190017536, 0.789672624785453, 0.402808177983388, 0.492040803655982, 0.882912602974102, 0.415280387736857, 0.603282547555864, 0.113552388502285, 0.839338998077437, 0.574409062275663, 0.59469018317759, 0.0792129875626415, 0.337749135680497, 0.249444715213031, 0.357683059060946, 0.560686685377732, 0.205485159764066, 0.517738219117746, 0.356066736858338, 0.575213237432763, 0.204983521252871, 0.320751714287326, 0.548385109053925, 0.00297991652041674, 0.841301759704947, 0.944254608126357, 0.0136947901919484, 0.981038260972127, 0.984887800645083, 0.818288646871224, 0.912047318648547, 0.663401561323553, 0.32243850082159, 0.950182125205174, 0.452226613881066, 0.0427293945103884, 0.199429222615436, 0.207023217342794, 0.339213469764218, 0.125580113381147, 0.632216554600745, 0.665811332175508, 0.787949427729473, 0.40842007054016, 0.195793261518702, 0.686935421079397, 0.107897110283375, 0.623920921934769, 0.0743551605846733, 0.683160332264379, 0.753325073281303, 0.00472963717766106, 0.0955639253370464, 0.298760182224214, 0.131604062393308, 0.770932401996106, 0.70556173636578, 0.549382609780878, 0.100654212059453, 0.751124553149566, 0.520781392697245, 0.0846729376353323, 0.659375834045932, 0.283012792933732, 0.538408106192946, 0.593145299004391, 0.352760220179334, 0.411563335452229, 0.310060786549002, 0.944449327187613, 0.10515916114673, 0.583528950344771, 0.457482134457678, 0.0274235815741122, 0.122501741629094, 0.774305870756507, 0.977638724958524, 0.591862949542701, 0.180453789187595, 0.988440885441378, 0.578901817556471, 0.195425867568702, 0.75391267426312, 0.868517672410235, 0.154747667722404, 0.61089889658615, 0.0941362152807415, 0.119190794182941, 0.451597329461947, 0.714477586327121, 0.463533287169412, 0.127074834192172, 0.351436968427151, 0.581865913933143, 0.562359639676288, 0.945521359331906, 0.952316353563219, 0.402391623705626, 0.457315566716716, 0.218097717035562, 0.675314189400524, 0.691628420958295, 0.214284206042066, 0.60111265280284, 0.312624556012452, 0.383372225100175, 0.716730079380795, 0.539852875517681, 0.108836917439476, 0.0478576757013798, 0.114313907921314, 0.225960201816633, 0.20469997311011, 0.219574862392619, 0.51736683677882, 0.701591759454459, 0.621795326704159, 0.780488878721371, 0.904020769987255, 0.962906744331121, 0.819297080626711, 0.520253876456991, 0.751169484108686, 0.241246284684166, 0.879281983012334, 0.592125033959746, 0.629448713967577, 0.460834025172517, 0.558868357911706, 0.945204924792051, 0.612336804624647, 0.763877178542316, 0.0880746934562922, 0.766706527676433, 0.216140698641539, 0.573912609601393, 0.838702959241346, 0.864720249082893, 0.0132389508653432, 0.661906033055857, 0.788639496546239, 0.815930024720728, 0.85209298087284, 0.769918238278478, 0.746297364355996, 0.303769884165376, 0.991051138145849, 0.947790045756847, 0.313885086448863, 0.494702912634239, 0.995048500131816, 0.223817711230367, 0.888250693446025, 0.267435152083635, 0.799045568332076, 0.497089291457087, 0.766665728762746, 0.897716783452779, 0.643443110166118, 0.570542548550293, 0.112623317865655, 0.587875941768289, 0.329850657144561, 0.503696311730891, 0.586600023088977, 0.479699927847832, 0.293491856660694, 0.669715109979734, 0.176418439485133, 0.64806009712629, 0.823010785738006, 0.917695899726823, 0.325642011361197, 0.945453931111842, 0.757537704659626, 0.862197360256687, 0.119183494942263, 0.184699904639274, 0.54240398388356, 0.435939426068217, 0.0724356269929558, 0.600686365971342, 0.499900706810877, 0.0573608493432403, 0.556534197647125, 0.11893312074244, 0.791977106360719, 0.617158843902871, 0.321729080518708, 0.120115455938503, 0.537801103200763, 0.727600121404976, 0.514765352709219, 0.388379360549152, 0.279221983859316, 0.204063391312957, 0.128110548248515, 0.857437104918063, 0.584097370272502, 0.721833146410063, 0.395581619581208, 0.833768550073728, 0.641163889318705, 0.6451605418697, 0.193209277000278, 0.0976289473474026, 0.179284179816023, 0.761480161454529, 0.494840135099366, 0.920421563321725, 0.646129658445716, 0.899439395172521, 0.0100232448894531, 0.637091716518626, 0.661029360024258, 0.988761197309941, 0.668709255987778, 0.268593549029902, 0.886470920406282, 0.645275991875678, 0.719777700258419, 0.238741755485535, 0.503916524816304, 0.306444169953465, 0.535812705755234, 0.878144236281514, 0.496176151791587, 0.564157212385908, 0.26975223259069, 0.250886202557012, 0.250578949227929, 0.0407194856088609, 0.186899994499981, 0.228765027131885, 0.651754448888823, 0.996928529348224, 0.135425538988784, 0.771570580545813, 0.841248871060088, 0.346957123139873, 0.112717882264405, 0.942783164558932, 0.0958205091301352, 0.687142799142748, 0.497039095265791, 0.174642578931525, 0.754032387863845, 0.248018547194079, 0.483546854695305, 0.812042643781751, 0.254188856342807, 0.898295036284253, 0.746629890985787, 0.308261223835871, 0.425615655723959, 0.611307506915182, 0.113459077430889, 0.345539406174794, 0.601398935541511, 0.721151690930128, 0.999025488039479, 0.512922743801028, 0.527555829845369, 0.977785886032507, 0.120036243926734, 0.433077562833205, 0.213218990480527, 0.916916372952983, 0.540808373363689, 0.525922448607162, 0.689642985351384, 0.543333060108125, 0.808610185282305, 0.575346032856032, 0.706563757034019, 0.717369175283238, 0.538707501254976, 0.196351790102199, 0.286341850180179, 0.640584216220304, 0.50354279531166, 0.223558027297258, 0.302394631085917, 0.4286102168262, 0.802115625236183, 0.229079304961488, 0.0234703093301505, 0.418002953985706, 0.483687452739105, 0.441213882761076, 0.497418868588284, 0.817127837333828, 0.737975085387006, 0.816587866749614, 0.361391683807597, 0.889178999466822, 0.452126294374466, 0.61843513417989, 0.932716809678823, 0.733726737555116, 0.780627983389422, 0.228670499287546, 0.213549250271171, 0.391368541866541, 0.589696795912459, 0.960442793788388, 0.423575647873804, 0.293472953373566, 0.718923375708982, 0.532499115448445, 0.35562106105499, 0.0973054114729166, 0.397503949701786, 0.526795492274687, 0.0361136316787452, 0.828915460500866, 0.633166905492544, 0.113619836745784, 0.0995861054398119, 0.19382669730112, 0.249115602811798, 0.520593663910404, 0.168483019806445, 0.193176351487637, 0.549088369123638, 0.10106193413958, 0.409682713449001, 0.372136724414304, 0.0662080263718963, 0.575420735403895, 0.403170882258564, 0.546471954090521, 0.779018131550401, 0.511760841589421, 0.883482729783282, 0.494599330704659, 0.253036069683731, 0.593815740430728, 0.889911918435246, 0.585185057250783, 0.23490095580928, 0.00530372792854905, 0.943676156224683, 0.0566373902838677, 0.888595525408164, 0.111225998261943, 0.911275469465181, 0.141761716688052, 0.688520044554025, 0.518743140855804, 0.772990550613031, 0.123625197215006, 0.7258410458453, 0.0411390902008861, 0.40936491987668, 0.671751619549468, 0.23510004975833, 0.86234272364527, 0.075170363066718, 0.857345885131508, 0.214257874293253, 0.46705341222696, 0.189096170244738, 0.0564745124429464, 0.896773590240628, 0.399305902887136, 0.922025207895786, 0.666010633111, 0.880088513717055, 0.260594850638881, 0.369925890816376, 0.945831426652148, 0.331352877430618, 0.17201925185509, 0.306876091053709, 0.87145282025449, 0.509524384280667, 0.0299883505795151, 0.113258351339027, 0.457362776622176, 0.703724470222369, 0.895382763585076, 0.195111000910401, 0.207648905226961, 0.172510000644252, 0.374085446819663, 0.973215428879485, 0.816590594127774, 0.815735967596993, 0.147526447661221, 0.202996098902076, 0.687684140633792, 0.658893079264089, 0.746211674064398, 0.217539184959605, 0.0377947324886918, 0.791612050263211, 0.848851593676955, 0.966748545411974, 0.926209100987762, 0.953244507312775, 0.325075415195897, 0.911368643864989, 0.356294434284791, 0.455728119704872, 0.85301089566201, 0.807395339012146, 0.510375774931163, 0.93307636748068, 0.584233234170824, 0.756082838401198, 0.142914057476446, 0.575494903838262, 0.115476104198024, 0.498747917823493, 0.425867939833552, 0.0142016501631588, 0.832725153071806, 0.605321697192267, 0.938236252870411, 0.343188217841089, 0.758399602025747, 0.805526421638206, 0.755987033247948, 0.57743874983862, 0.392967495834455, 0.718814110849053, 0.5285975697916, 0.448661683360115, 0.949267980176955, 0.298338768305257, 0.368053312646225, 0.935122460359707, 0.235074971336871, 0.342900226125494, 0.663329789415002, 0.260339515283704, 0.583988682134077, 0.384414691943675, 0.486597301671281, 0.852057240903378, 0.466257594991475, 0.220346237067133, 0.638610836816952, 0.408242274541408, 0.579871822614223, 0.899088418809697, 0.493008843623102, 0.181493946816772, 0.664959558052942, 0.263063790276647, 0.377947842003778, 0.854571215808392, 0.139875392429531, 0.485519677167758, 0.8767946572043, 0.867089342791587, 0.99714264459908, 0.553358307806775, 0.926510237390175, 0.647594016045332, 0.0324250445701182, 0.322340924292803, 0.616157486569136, 0.611021982273087, 0.610205311328173, 0.558478479972109, 0.723009368637577, 0.338565715122968, 0.437625793041661, 0.119364888640121, 0.463243185309693, 0.0634381910786033, 0.117582456907257, 0.552838415373117, 0.781930432422087, 0.488447251264006, 0.353684758301824, 0.589892990887165, 0.475381863536313, 0.867002913495526, 0.488301923731342, 0.312711394159123, 0.674997708294541, 0.0350598474033177, 0.842517815995961, 0.388702663360164, 0.912709114607424, 0.558871655026451, 0.139302268158644, 0.920758419902995, 0.503191696479917, 0.824417277239263, 0.247807351639494, 0.380398215726018, 0.866731214569882, 0.0224764510057867, 0.00900901528075337, 0.619516495382413, 0.0401400835253298, 0.353365032467991, 0.863583341473714, 0.25165240955539, 0.525249452563003, 0.861241003265604, 0.274258462944999, 0.432662300532684, 0.815361882792786, 0.197130168089643, 0.0264466984663159, 0.985924552427605, 0.276380046736449, 0.961070246528834, 0.809785472461954, 0.55315769976005, 0.901994448155165, 0.115694968728349, 0.831107018748298, 0.036921443650499, 0.731991172302514, 0.790078349178657, 0.684765264391899, 0.175432358169928, 0.0508316506166011, 0.484106645220891, 0.863742682151496, 0.532323576509953, 0.0218174245674163, 0.982238253811374, 0.183637833921239, 0.000521493144333363, 0.445002931635827, 0.410807237261906, 0.542195119196549, 0.829936088062823, 0.968061733059585, 0.947781693655998, 0.147042684489861, 0.889401260763407, 0.820332010043785, 0.283112226752564, 0.105021384079009, 0.809806478675455, 0.724546989658847, 0.185911279171705, 0.226517034461722, 0.0614388193935156, 0.890772240236402, 0.129657773068175, 0.268135221675038, 0.598965923534706, 0.433622139506042, 0.631700177444145, 0.690776854520664, 0.952111603459343, 0.612607039744034, 0.897398667177185, 0.202449552481994, 0.51349010085687, 0.243563815252855, 0.232952755177394, 0.0958858085796237, 0.672022800659761, 0.857687646057457, 0.617909194203094, 0.868237990653142, 0.627980643650517, 0.554121889872476, 0.751004436518997, 0.217555851908401, 0.0531074849423021, 0.919539601542056, 0.384487601928413, 0.839454363565892, 0.641728626564145, 0.928542794194072, 0.0655185848008841, 0.423124228138477, 0.0809074528515339, 0.285919656278566, 0.0758803621865809, 0.854872543597594, 0.935287588043138, 0.656797407427803, 0.139227515319362, 0.377164341043681, 0.624391376040876, 0.882313327863812, 0.295059587573633, 0.503154906909913, 0.0499138499144465, 0.154868601821363, 0.749411268858239, 0.503670096164569, 0.825444026151672, 0.957327157258987, 0.102301482809708, 0.710190040525049, 0.352105807512999, 0.8708852941636, 0.205306294606999, 0.584750112146139, 0.148888568626717, 0.623978942167014, 0.770735974656418, 0.289947077166289, 0.148377734003589, 0.79654089990072, 0.270381319103763, 0.894084709696472, 0.978985350346193, 0.505612606415525, 0.74497089185752, 0.632727886084467, 0.557267564116046, 0.0279436658602208, 0.478339898400009, 0.56932564266026, 0.769641816848889, 0.492598346434534, 0.419692617375404, 0.772737013641745, 0.101124224485829, 0.232093322789297, 0.77777765924111, 0.345272416481748, 0.185002156533301, 0.798082896042615, 0.0796995162963867, 0.841147697996348, 0.695000291569158, 0.361129351425916, 0.227007374167442, 0.759570384863764, 0.751393652055413, 0.181262521306053, 0.653355882270262, 0.794331038137898, 0.906257794937119, 0.72985321492888, 0.429292797576636, 0.677500877762213, 0.0171030496712774, 0.651718146400526, 0.623287261463702, 0.321811289060861, 0.107134556863457, 0.694210693240166, 0.50469172000885, 0.404249225743115, 0.6205773872789, 0.920515005476773, 0.611277975142002, 0.594512095442042, 0.639725558692589, 0.671402819221839, 0.460087034385651, 0.953793410910293, 0.271082833409309, 0.895330559229478, 0.435948475729674, 0.116468336200342, 0.0739294711966068, 0.335408284561709, 0.85956027591601, 0.196365308016539, 0.296481077326462, 0.832491416018456, 0.620470809284598, 0.758886501658708, 0.953366628149524, 0.287902995012701, 0.60188511852175, 0.565231148852035, 0.916367873549461, 0.504073035204783, 0.976277461042628, 0.809064597357064, 0.466050931019709, 0.812877719989046, 0.288722644094378, 0.370032444829121, 0.694573185173795, 0.45900951186195, 0.906701085856184, 0.538248464232311, 0.252976593794301, 0.416728546377271, 0.0682451508473605, 0.839265105081722, 0.875755267916247, 0.488286844687536, 0.120410901959985, 0.54960914934054, 0.343640452250838, 0.324930424336344, 0.314384143101051, 0.0684066235553473, 0.829648860730231, 0.147507640533149, 0.856647948967293, 0.39628217369318, 0.548444728832692, 0.366546351462603, 0.304108169861138, 0.834494072012603, 0.376272874884307, 0.0877235203515738, 0.70465329894796, 0.219467185204849, 0.188644569832832, 0.0579102269839495, 0.00529035506770015, 0.684364063665271, 0.63658180902712, 0.674473216524348, 0.857812949223444, 0.770939795766026, 0.574451088206843, 0.698381746187806, 0.356688797473907, 0.97295853192918, 0.815675121499225, 0.202863920945674, 0.71144665684551, 0.480627194978297, 0.201661884319037, 0.293050199048594, 0.625633246032521, 0.163198349066079, 0.735878193983808, 0.134046335471794, 0.735193325439468, 0.710159965790808, 0.0419055114034563, 0.65104173286818, 0.23037480073981, 0.32933792239055, 0.264086165465415, 0.478332246653736, 0.118887348100543, 0.0570161824580282, 0.839999163756147, 0.169259962858632, 0.31541115581058, 0.0644873783458024, 0.565827093552798, 0.752067785011604, 0.617873232578859, 0.634147083153948, 0.987623565364629, 0.689049502136186, 0.810807170346379, 0.848693719832227, 0.633458721684292, 0.751307700993493, 0.0343727848958224, 0.68760229786858, 0.683163266628981, 0.374537569936365, 0.947080082958564, 0.279151720460504, 0.96038145525381, 0.949563813395798, 0.296568828634918, 0.719787407666445, 0.318903286708519, 0.0391541286371648, 0.450560038210824, 0.556174904806539, 0.497451805742458, 0.0233313057105988, 0.843052873155102, 0.38470096513629, 0.147318796487525, 0.722046926151961, 0.188216614536941, 0.973652645014226, 0.975906708277762, 0.856179665774107, 0.918974731350318, 0.011147472076118, 0.493533584056422, 0.951549293473363, 0.59626767411828, 0.585728566162288, 0.172967196442187, 0.256266971584409, 0.693231307901442, 0.310357791837305, 0.952485613059253, 0.767077138414606, 0.907794231083244, 0.301295108161867, 0.720165741862729, 0.474845174700022, 0.403149889083579, 0.175586240366101, 0.881843127077445, 0.0670842591207474, 0.950717016821727, 0.429001931799576, 0.833391501335427, 0.259404621319845, 0.90695969783701, 0.628544099628925, 0.857553884387016, 0.693010705756024), actual = c(0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"