Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.6027690        0.6027690 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4441084        0.4441084        0.5164676 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.824771264567971, 0.718936000484973, 0.666654666885734, 0.575622686417773, 0.596924874000251, 0.820272714132443, 0.494691514177248, 0.149794179713354, 0.183581040007994, 0.627761522307992, 0.741946908878163, 0.563394573284313, 0.251457558246329, 0.85915955575183, 0.616116123041138, 0.731263901107013, 0.172575457254425, 0.279061476932839, 0.192614273400977, 0.0766854975372553, 0.460923263803124, 0.209457731805742, 0.631656613433734, 0.261765741277486, 0.382532855728641, 0.485590194119141, 0.205696295714006, 0.660619199974462, 0.740409811260179, 0.442751368973404, 0.235637322533876, 0.670995727879927, 0.239686702145264, 0.0300418015103787, 0.243184556486085, 0.162229184061289, 0.516825861297548, 0.822181512136012, 0.0336261901538819, 0.444106662413105, 0.28205911652185, 0.0637623006477952, 0.623041605111212, 0.553701594704762, 0.645625890698284, 0.668124626623467, 0.149745028233156, 0.625227500684559, 0.35585539531894, 0.708970215637237, 0.593893483746797, 0.791769586037844, 0.452755525009707, 0.941230179741979, 0.450379228917882, 0.324036604026332, 0.293843549443409, 0.0726951591204852, 0.737914126133546, 0.330396917415783, 0.179461054736748, 0.940860922681168, 0.47388044022955, 0.409306450281292, 0.930373226990923, 0.744354663416743, 0.836902656359598, 0.213609730591998, 0.850881638005376, 0.194669369608164, 0.704357113456354, 0.0983492962550372, 0.0771658869925886, 0.17130342614837, 0.848151239100844, 0.163918594131246, 0.69784017605707, 0.251450164942071, 0.93489854503423, 0.466507159173489, 0.77478471910581, 0.274661565897986, 0.894239761633798, 0.297822097782046, 0.313801730750129, 0.568763646064326, 0.0290943044237792, 0.291161750210449, 0.240927652223036, 0.983584750909358, 0.379165584221482, 0.0867144828662276, 0.170662680640817, 0.798580311471596, 0.351485240040347, 0.877698690863326, 0.0560214477591217, 0.202751860721037, 0.323614862980321, 0.394984483951703, 0.296809893799946, 0.296126884873956, 0.875368320383132, 0.664253985043615, 0.467001234879717, 0.562673431821167, 0.792324557900429, 0.110300995409489, 0.255435423459858, 0.0345684383064508, 0.940515369409695, 0.219097990076989, 0.980503253638744, 0.734168632188812, 0.595633269986138, 0.110818400047719, 0.432558752363548, 0.154325139475986, 0.245798397576436, 0.471648362232372, 0.618716469500214, 0.796292160404846, 0.12644550204277, 0.481281747575849, 0.816251837881282, 0.610907441005111, 0.0816455308813602, 0.736061305273324, 0.635592364706099, 0.0246295176912099, 0.941663721576333, 0.139741056598723, 0.722662897547707, 0.807548079406843, 0.647352231666446, 0.776619729585946, 0.526571501977742, 0.580024812137708, 0.283553290180862, 0.112109756795689, 0.785215468145907, 0.65817755064927, 0.0142207550816238, 0.264224804239348, 0.210426481440663, 0.93431334104389, 0.906383288558573, 0.39127629972063, 0.808145960792899, 0.0671114344149828, 0.99158352590166, 0.570847088005394, 0.897709532175213, 0.310333049623296, 0.452284236904234, 0.86521956580691, 0.843365569133312, 0.505875814240426, 0.187992450082675, 0.686913463287055, 0.176459396490827, 0.952279447810724, 0.46889400924556, 0.722033390309662, 0.426503991009668, 0.256123375147581, 0.0397821187507361, 0.636786643182859, 0.435055249836296, 0.525491480249912, 0.345929237315431, 0.241693513002247, 0.806126557523385, 0.50065570208244, 0.723029171582311, 0.0808257299941033, 0.815540651325136, 0.686042422661558, 0.70189632801339, 0.191616172902286, 0.98334415955469, 0.66118728206493, 0.0863083079457283, 0.937149713747203, 0.926214793697, 0.0964555363170803, 0.190984998829663, 0.611012087203562, 0.81575130764395, 0.162105570547283, 0.332258585607633, 0.883639416191727, 0.672374834306538, 0.641664676601067, 0.838573640678078, 0.584187166532502, 0.0127813413273543, 0.463175580138341, 0.762772152898833, 0.292622189735994, 0.648603944573551, 0.768090056721121, 0.756680060643703, 0.805074823554605, 0.170359572861344, 0.17346960073337, 0.799371039262041, 0.501195392105728, 0.645900538191199, 0.447504214243963, 0.730641184607521, 0.407237522071227, 0.203720451565459, 0.233971780631691, 0.695063104154542, 0.841597324237227, 0.712468727957457, 0.0240680652204901, 0.895364600466564, 0.264950510114431, 0.382993525825441, 0.0373828944284469, 0.508128088666126, 0.0380508066155016, 0.0461080465465784, 0.452352892840281, 0.431426614057273, 0.0513589172624052, 0.431686982279643, 0.551177945220843, 0.00505313719622791, 0.859244011808187, 0.191701216856018, 0.0906660230830312, 0.701330081326887, 0.547561481129378, 0.741602865280584, 0.886665574274957, 0.146916619036347, 0.338127824245021, 0.182195161003619, 0.530169468605891, 0.158289678394794, 0.123172220308334, 0.709440044360235, 0.372042735572904, 0.254498188151047, 0.0893719710875303, 0.329294410999864, 0.609258828684688, 0.0136342204641551, 0.17349356552586, 0.772242511389777, 0.217354471795261, 0.187566057080403, 0.751346078002825, 0.965565438382328, 0.0696682066190988, 0.729363309685141, 0.78998903837055, 0.420518935425207, 0.815906976582482, 0.78228775691241, 0.505278728669509, 0.926934377290308, 0.216982994694263, 0.192006500903517, 0.263751879101619, 0.604846769478172, 0.609039792558178, 0.753577502677217, 0.230160685721785, 0.421469150111079, 0.0255994047038257, 0.700631093932316, 0.801122306846082, 0.760580968344584, 0.236533182673156, 0.987983875442296, 0.483131326036528, 0.332885856274515, 0.830648157047108, 0.135924568399787, 0.339112795423716, 0.287665097974241, 0.747304418357089, 0.582156239310279, 0.538984139682725, 0.717458726139739, 0.610436334973201, 0.689038194948807, 0.760166492778808, 0.0611032729502767, 0.0278216968290508, 0.252923734253272, 0.484856865834445, 0.914653236279264, 0.546744090737775, 0.402283917879686, 0.279202758567408, 0.921934329438955, 0.862215158995241, 0.995980951702222, 0.397110996302217, 0.492396759334952, 0.462913925526664, 0.357416299404576, 0.518264959100634, 0.639110013609752, 0.879529035650194, 0.716901185922325, 0.00954156927764416, 0.193895664531738, 0.0537499666679651, 0.465930325910449, 0.807100086240098, 0.491349162301049, 0.524735179264098, 0.519186273217201, 0.515137293376029, 0.58020231127739, 0.0376131718512625, 0.117887587286532, 0.624202495673671, 0.121990120504051, 0.625930276699364, 0.912207449087873, 0.71333015547134, 0.340484134154394, 0.247338552726433, 0.107896376168355, 0.809155691647902, 0.987307444680482, 0.682791230734438, 0.0590665913186967, 0.373393616406247, 0.905543752247468, 0.584813671885058, 0.452520343009382, 0.518518076743931, 0.751227585831657, 0.144244478084147, 0.00268885004334152, 0.828638098435476, 0.418440179200843, 0.347161531914026, 0.00212395936250687, 0.0907383714802563, 0.729160117916763, 0.667249280493706, 0.114973631221801, 0.738348940154538, 0.96328422264196, 0.6071971738711, 0.484251231886446, 0.559474302222952, 0.598262784071267, 0.105424162000418, 0.0805931966751814, 0.929509358247742, 0.442292690742761, 0.419204891659319, 0.0302667832002044, 0.457901014247909, 0.94615814788267, 0.315685638226569, 0.0148113125469536, 0.75086012436077, 0.674217924242839, 0.856566885020584, 0.129400139907375, 0.727788008749485, 0.10420525399968, 0.730795185780153, 0.120729435235262, 0.79815784865059, 0.795650297077373, 0.546580748865381, 0.506032409379259, 0.224367255112156, 0.890003411099315, 0.503147449111566, 0.65985733945854, 0.374138575512916, 0.530034067574888, 0.117838995298371, 0.359407304320484, 0.778957592789084, 0.16766429389827, 0.0155190583318472, 0.625639891717583, 0.712524248752743, 0.704051965149119, 0.317668030969799, 0.726042175898328, 0.0971201111096889, 0.874241233803332, 0.590115470811725, 0.359142533736303, 0.221414111787453, 0.0510928195435554, 0.255088117904961, 0.977572857635096, 0.212956645060331, 0.68627268425189, 0.474946882110089, 0.188274447107688, 0.789765371475369, 0.0779071170836687, 0.267795259132981, 0.948060748400167, 0.123128736391664, 0.463508110959083, 0.356021360959858, 0.622341142268851, 0.925689955707639, 0.500960114179179, 0.961585939629003, 0.224894761806354, 0.795181419933215, 0.884126741904765, 0.501089384779334, 0.0592526646796614, 0.64505125163123, 0.248031246010214, 0.410786150256172, 0.530494695995003, 0.202899115625769, 0.221288125496358, 0.241342982044443, 0.831961089046672, 0.171581198461354, 0.33207874558866, 0.899408578872681, 0.0830995738506317, 0.349576559849083, 0.43226898717694, 0.332982504041865, 0.469379094662145, 0.601414912380278, 0.584676663391292, 0.106774196960032, 0.508666380075738, 0.531273935222998, 0.0350039447657764, 0.925673409132287, 0.0780534925870597, 0.971333711175248, 0.543017400894314, 0.0147708628792316, 0.0634101275354624, 0.68271583202295, 0.832923491951078, 0.0876408026088029, 0.00498636742122471, 0.984107969794422, 0.448882165597752, 0.975220675114542, 0.877648829016834, 0.556675853207707, 0.848963879281655, 0.123478940920904, 0.701096304459497, 0.171157752396539, 0.515697689261287, 0.925634126644582, 0.856923238839954, 0.671977050602436, 0.736757631180808, 0.262559655820951, 0.991734765702859, 0.916830761358142, 0.449822401395068, 0.570863120723516, 0.685730592347682, 0.613358356757089, 0.945124663412571, 0.358032362302765, 0.888928011991084, 0.089096125215292, 0.363965378841385, 0.316066493745893, 0.491683210246265, 0.214264696929604, 0.279733301606029, 0.734839731594548, 0.215336952591315, 0.28787499293685, 0.774521747138351, 0.282549346331507, 0.49914619885385, 0.449104457860813, 0.959794507129118, 0.858415466500446, 0.635676198173314, 0.119999132817611, 0.269661530852318, 0.226316424319521, 0.393183502368629, 0.883807291509584, 0.690719051053748, 0.0630655330605805, 0.954608629457653, 0.174139938782901, 0.470182423014194, 0.739026481984183, 0.69698700052686, 0.156986138783395, 0.144429031992331, 0.314716866705567, 0.740309846587479, 0.861049060709774, 0.77610981091857, 0.264694353099912, 0.556595139671117, 0.0055964644998312, 0.4288501215633, 0.571973441634327, 0.480668774107471, 0.634791076881811, 0.680614370154217, 0.0848000978585333, 0.212020603008568, 0.617635992821306, 0.584012474166229, 0.299639338627458, 0.773116559255868, 0.437435871688649, 0.532517238752916, 0.145412820857018, 0.108206506818533, 0.214920466067269, 0.983177911024541, 0.191351871006191, 0.5645117003005, 0.551580948987976, 0.773898189188913, 0.785899077774957, 0.0658349408768117, 0.380137583706528, 0.526495964964852, 0.193086331011727, 0.11093011777848, 0.0126135968603194, 0.241825076984242, 0.631248497171327, 0.659072452224791, 0.260145251173526, 0.608247208409011, 0.600010201102123, 0.801620514597744, 0.762373590609059, 0.584703062195331, 0.329818622441962, 0.366647427668795, 0.818965746788308, 0.485053981421515, 0.313167269807309, 0.681932976702228, 0.139873019652441, 0.44719795207493, 0.337586681824178, 0.149063556222245, 0.48133390257135, 0.412251525325701, 0.904471489135176, 0.146514319581911, 0.493734831223264, 0.9546528251376, 0.364228665130213, 0.812903794692829, 0.734928647987545, 0.936726429732516, 0.575245101703331, 0.496195531683043, 0.612019500695169, 0.893243710044771, 0.556679700966924, 0.163180232513696, 0.66346873389557, 0.652996693504974, 0.0446061752736568, 0.571104389615357, 0.743415567558259, 0.250709761399776, 0.262960960622877, 0.0420914643909782, 0.00511147500947118, 0.0652183955535293, 0.93680465198122, 0.970384446205571, 0.997809371445328, 0.20207015867345, 0.988034747075289, 0.397004878614098, 0.0305343861691654, 0.805125090759248, 0.0785226910375059, 0.98971915920265, 0.13965718797408, 0.920791264623404, 0.461736519122496, 0.42506449110806, 0.0315607476513833, 0.885273406514898, 0.247258242452517, 0.944390145828947, 0.945690524065867, 0.868238318478689, 0.130857602693141, 0.0700093847699463, 0.787831831723452, 0.138473199680448, 0.557846704032272, 0.956589923938736, 0.190340608125553, 0.153673894004896, 0.0901399729773402, 0.259036167990416, 0.410364119336009, 0.969865903258324, 0.323504153871909, 0.30523042101413, 0.889951987657696, 0.904877473600209, 0.648107834625989, 0.462839375250041, 0.915163356345147, 0.710309784160927, 0.0154981454834342, 0.289611693006009, 0.7060937366914, 0.66743867052719, 0.103287275880575, 0.715235069394112, 0.109405822353438, 0.618156791431829, 0.391767568420619, 0.606436570640653, 0.933357923524454, 0.502445004647598, 0.518138556275517, 0.813364834990352, 0.611351819243282, 0.787540263030678, 0.301847473718226, 0.775382910389453, 0.716174351517111, 0.032919755904004, 0.0871018678881228, 0.998561405809596, 0.562034934526309, 0.662325581070036, 0.576497997622937, 0.798398091923445, 0.184469709871337, 0.831635089125484, 0.49339320580475, 0.101878381101415, 0.0150967631489038, 0.256251860177144, 0.871511749923229, 0.13406376237981, 0.652731795795262, 0.500344611238688, 0.41087045148015, 0.812270161230117, 0.121025298256427, 0.541689620353281, 0.415945035638288, 0.819640296977013, 0.311054522870108, 0.966000305488706, 0.398189390776679, 0.213374230777845, 0.552637712564319, 0.270184121560305, 0.183748679468408, 0.252157582668588, 0.582948826253414, 0.947500988375396, 0.0837032396811992, 0.128317124210298, 0.209531849250197, 0.672847996233031, 0.127921439008787, 0.170996434288099, 0.148124249884859, 0.394732360728085, 0.902311970479786, 0.646847952855751, 0.991821171948686, 0.723596374969929, 0.867235596291721, 0.637170328525826, 0.411631393246353, 0.289709491888061, 0.63887818576768, 0.42295310809277, 0.908781291218475, 0.184189453022555, 0.861503544263542, 0.639976593200117, 0.249236982082948, 0.347279290435836, 0.986195666715503, 0.396256688283756, 0.852798709413037, 0.906256948364899, 0.793555930722505, 0.0594097934663296, 0.255281665129587, 0.810391464270651, 0.711776601849124, 0.388832106953487, 0.373878179118037, 0.203266682336107, 0.182284256909043, 0.553600772051141, 0.053946757921949, 0.782245816430077, 0.104212658014148, 0.706088761566207, 0.159452563384548, 0.375798751600087, 0.521455079782754, 0.519127577310428, 0.256301766028628, 0.194454740965739, 0.527660834370181, 0.270980712492019, 0.135536247398704, 0.215016240719706, 0.902634915430099, 0.336984898196533, 0.228357906220481, 0.155899995472282, 0.795421345159411, 0.106070867273957, 0.488237111829221, 0.735832249745727, 0.887391510419548, 0.57264468120411, 0.490770452655852, 0.930611928226426, 0.994033084716648, 0.165363400010392, 0.566007263492793, 0.754315446363762, 0.176530310418457, 0.537661765702069, 0.0955032829660922, 0.219992802944034, 0.878725259099156, 0.20337524311617, 0.486850776243955, 0.20285118627362, 0.291131727164611, 0.403428890975192, 0.943818469997495, 0.718112479429692, 0.446541203884408, 0.898883201414719, 0.463032758561894, 0.715673693688586, 0.235957906348631, 0.226966558722779, 0.202641007024795, 0.0848876875825226, 0.165232247672975, 0.772011881228536, 0.820334793766961, 0.946428820258006, 0.617115644970909, 0.952477916842327, 0.1024092908483, 0.632295222021639, 0.0675005021039397, 0.930302399443462, 0.779513839399442, 0.325330410618335, 0.825761157087982, 0.420784490182996, 0.806636044522747, 0.660818375647068, 0.060815084958449, 0.90809833840467, 0.197735025547445, 0.691955782240257, 0.111342433141544, 0.0227347032632679, 0.375477170338854, 0.877345509361476, 0.826432584552094, 0.877489971229807, 0.239193216431886, 0.42240475746803, 0.264530721353367, 0.647409649565816, 0.77266588085331, 0.656319710193202, 0.396119062788785, 0.108994770795107, 0.257141658104956, 0.190082406625152, 0.404925363603979, 0.595103727886453, 0.124356514541432, 0.503818565048277, 0.33390673622489, 0.867633118061349, 0.210930602392182, 0.139030238380656, 0.066294654738158, 0.438192630186677, 0.768269685329869, 0.419340158812702, 0.867697000270709, 0.44998786249198, 0.606449472717941, 0.849611624609679, 0.351091786753386, 0.996905938722193, 0.281284674070776, 0.761970733525231, 0.923036370193586, 0.119133722269908, 0.251889027189463, 0.197336540091783, 0.70028204889968, 0.40910426992923, 0.833326433552429, 0.740042181685567, 0.285490741487592, 0.650204899720848, 0.122253195149824, 0.31566477403976, 0.571769438683987, 0.652382601052523, 0.707180769415572, 0.29168094182387, 0.796269299695268, 0.698740328429267, 0.785160395782441, 0.391942461021245, 0.475259814877063, 0.972144973929971, 0.063570344587788, 0.0226196346338838, 0.919364220229909, 0.82247591461055, 0.670614618808031, 0.960963369114324, 0.568288403796032, 0.836982831824571, 0.349479910451919, 0.251850143773481, 0.098981388611719, 0.636654434492812, 0.649438938125968, 0.155929069500417, 0.853504854720086, 0.306396714411676, 0.852105489233509, 0.671041492605582, 0.987706342246383, 0.309341402025893, 0.28401656774804, 0.451531453523785, 0.967930323909968, 0.687140505295247, 0.95512868068181, 0.220566496718675, 0.219748801086098, 0.968462640885264, 0.428933765739202, 0.0711779771372676, 0.278053235029802, 0.59211689629592, 0.466054647229612, 0.763482461450621, 0.532573523232713, 0.782611379865557, 0.581298647681251, 0.663221408380195, 0.306829864624888, 0.259059462463483, 0.0424076961353421, 0.643724079243839, 0.0571958660148084, 0.334548001177609, 0.168896910734475, 0.376904378179461, 0.265353628667071, 0.608121579047292, 0.0459480634890497, 0.515790397068486, 0.661815915955231, 0.101619591470808, 0.215670651057735, 0.0348263003397733, 0.921501904027537, 0.404109805589542, 0.762086795875803, 0.294733048649505, 0.199628825066611, 0.797340049641207, 0.430939809186384, 0.789505314547569, 0.856218158034608, 0.879441627301276, 0.130963160423562, 0.57017164491117, 0.421632844023407, 0.861646441044286, 0.60764137818478, 0.38039801036939, 0.242094671586528, 0.181007089326158, 0.719041324220598, 0.696941931266338, 0.307648944901302, 0.182559455279261, 0.356540864799172, 0.159962514415383, 0.0508843471761793, 0.415132847381756, 0.699638952966779, 0.287870833417401, 0.268467576242983, 0.781565148383379, 0.986708451295272, 0.0534083929378539, 0.397444114321843, 0.179311614716426, 0.876960037043318, 0.161137120099738, 0.865649664308876, 0.0766957968007773, 0.162297986447811, 0.716788361314684, 0.438441255595535, 0.40724507975392, 0.0649832922499627, 0.573188708396628, 0.634167420910671, 0.175428931834176, 0.85505076055415, 0.952599012991413, 0.69752989592962, 0.903520161285996, 0.661413851892576, 0.974937899736688, 0.688473944552243, 0.985147800296545, 0.794867928838357, 0.624671389814466, 0.978389851748943, 0.33079384220764, 0.796579418936744, 0.795120063936338, 0.727084547979757, 0.834938251413405, 0.674933781847358, 0.97145773540251, 0.780059685232118, 0.766714092576876, 0.379313023760915, 0.863265313673764, 0.367637463379651, 0.824961442267522, 0.225108146900311, 0.360792052466422, 0.827513994183391, 0.748308031121269, 0.488328010775149, 0.944056301610544, 0.927345101023093, 0.026889304863289, 0.725983050419018, 0.319640417583287, 0.87690194346942, 0.0119385616853833, 0.519442604854703, 0.173930167686194, 0.0833209287375212, 0.899747900431976, 0.771823808085173, 0.627243685070425, 0.300119707360864, 0.480784270213917, 0.970560310641304, 0.969376603839919, 0.0315206255763769, 0.647938752314076, 0.514839506940916, 0.0859956115018576, 0.330393012147397, 0.347994392737746, 0.417634805897251), actual = c(1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.109667072072625, 0.150022869929671, 0.867363044992089, 0.365523437736556, 0.823005464393646, 0.925906253745779, 0.95049539860338, 0.227535231038928, 0.641706431983039, 0.246792632853612, 0.241521178977564, 0.551649740431458, 0.0783193216193467, 0.862484233919531, 0.446303833276033, 0.689076020848006, 0.478076964151114, 0.971055993810296, 0.860642838990316, 0.989737002877519, 0.207983721513301, 0.86692179297097, 0.298485128907487, 0.878588026855141, 0.524494629818946, 0.757696821819991, 0.440107710193843, 0.766043413430452, 0.17647660872899, 0.294077965430915, 0.582802870078012, 0.999462397769094, 0.683681265451014, 0.161646307213232, 0.733953914605081, 0.405395512701944, 0.122131426353008, 0.925793293863535, 0.430764693534002, 0.353655216982588, 0.627707114210352, 0.34612259734422, 0.907437706831843, 0.874128555413336, 0.757416813401505, 0.832420841325074, 0.951613865094259, 0.098270149435848, 0.794207526138052, 0.809242694405839, 0.465195565717295, 0.282115692272782, 0.23750318121165, 0.998178230132908, 0.480113485595211, 0.140105778817087, 0.662660004803911, 0.616549248807132, 0.469378652516752, 0.962190485326573, 0.0500855376012623, 0.0898176922928542, 0.0402815355919302, 0.923255620058626, 0.597831372404471, 0.222505503334105, 0.114453875459731, 0.933962368639186, 0.72499511227943, 0.483861312037334, 0.885775482514873, 0.854363641468808, 0.190573030384257, 0.153498416999355, 0.174447417259216, 0.268485974520445, 0.588450798299164, 0.676593505777419, 0.804220300167799, 0.0597061454318464, 0.158320290502161, 0.594035255722702, 0.0722001360263675, 0.747924353927374, 0.162500451086089, 0.649248103611171, 0.224586763884872, 0.384577022399753, 0.219243143219501, 0.366169601911679, 0.901270658941939, 0.842482642503455, 0.733042309759185, 0.913044886663556, 0.402357026701793, 0.786719582276419, 0.271170767722651, 0.262715886346996, 0.710767977172509, 0.872660223627463, 0.330166239989921, 0.170899882912636, 0.668215128127486, 0.0688265657518059, 0.702483865432441, 0.850874488707632, 0.4577669093851, 0.293343745637685, 0.524134995648637, 0.120962776942179, 0.849613490747288, 0.888405429199338, 0.886121701681986, 0.852920397883281, 0.343848610995337, 0.355995862977579, 0.691739693749696, 0.80410789581947, 0.110372453695163, 0.985520168207586, 0.442853901069611, 0.0923460905905813, 0.278462872141972, 0.835905041079968, 0.623517729807645, 0.477974325651303, 0.72955536050722, 0.693357537733391, 0.91105273226276, 0.627556208753958, 0.161456034285948, 0.896078627556562, 0.784761386457831, 0.564903419697657, 0.149780607316643, 0.164796551223844, 0.436228736536577, 0.611836343538016, 0.66486417199485, 0.147008516592905, 0.356393214548007, 0.156860939227045, 0.0922132842242718, 0.187495566671714, 0.512302513699979, 0.0262695469427854, 0.0124260229058564, 0.0184870192315429, 0.547088890336454, 0.799677530536428, 0.305811529280618, 0.81158648733981, 0.270334733882919, 0.857566096587107, 0.961577519308776, 0.406224662903696, 0.979597543133423, 0.0229199803434312, 0.00982948672026396, 0.315407502232119, 0.129518431145698, 0.991771390195936, 0.0934990257956088, 0.986087072407827, 0.21930259023793, 0.685590751469135, 0.427011677064002, 0.578511378495023, 0.912479538004845, 0.0424499090295285, 0.277118241880089, 0.817893386352807, 0.177154831355438, 0.962262810673565, 0.895115157123655, 0.41902079177089, 0.659336341079324, 0.963662644382566, 0.896982758305967, 0.461956725921482, 0.594715020386502, 0.814733788138255, 0.417613492114469, 0.0812358160037547, 0.589627934852615, 0.294775901595131, 0.750523348804563, 0.381054576020688, 0.54738199384883, 0.35224341484718, 0.271301930071786, 0.367047695675865, 0.824672220274806, 0.317966796224937, 0.658410158706829, 0.117490106029436, 0.33421004540287, 0.668970587896183, 0.0104689300060272, 0.794678258476779, 0.556773749645799, 0.469502530060709, 0.581461673835292, 0.95566736231558, 0.097078200429678, 0.719497533515096, 0.110791109502316, 0.61244131764397, 0.356419375631958, 0.0953427688218653, 0.171399262268096, 0.979538178769872, 0.655273298500106, 0.17359761451371, 0.851698587182909, 0.254736374830827, 0.331629713298753, 0.144127778010443, 0.561318447813392, 0.487463199067861, 0.86797514767386, 0.175753314513713, 0.878712729783729, 0.646534659201279, 0.0902475928887725, 0.120527430903167, 0.194800483062863, 0.391071223421022, 0.48181113274768, 0.78820900968276, 0.225299844983965, 0.74446295038797, 0.754165123449638, 0.133026677649468, 0.0555337029509246, 0.189926340477541, 0.236546742031351, 0.640964877093211, 0.0387421550694853, 0.844431640813127, 0.616054810816422, 0.754775886423886, 0.45017614797689, 0.606608827365562, 0.736630781553686, 0.0570492024999112, 0.943098054034635, 0.203163350699469, 0.504082355415449, 0.709734508534893, 0.489808932179585, 0.460303504019976, 0.0669839426409453, 0.858487067976967, 0.229686401551589, 0.121411753818393, 0.561357847647741, 0.420612602028996, 0.00747680198401213, 0.808328705141321, 0.30495874513872, 0.843788949307054, 0.917526209261268, 0.846740514505655, 0.388344288105145, 0.266680735861883, 0.974349129712209, 0.0834432269912213, 0.726827314822003, 0.836410858202726, 0.0278509058989584, 0.585968083236367, 0.349909334909171, 0.986562005942687, 0.660959063796327, 0.404675864381716, 0.894665361847728, 0.477060268400237, 0.495675660902634, 0.0561113904695958, 0.51590840308927, 0.81217406084761, 0.503920986782759, 0.117435271851718, 0.0182422955986112, 0.598893607733771, 0.230664947535843, 0.664953037397936, 0.47971245390363, 0.417806069599465, 0.670967352110893, 0.220384675776586, 0.350398745620623, 0.184912732569501, 0.229994659777731, 0.763527768896893, 0.389794447924942, 0.725553218508139, 0.0498288199305534, 0.804245341103524, 0.651788460323587, 0.714377964148298, 0.424357743002474, 0.87554584746249, 0.968068164540455, 0.317608451237902, 0.0588226208928972, 0.742407559650019, 0.321332838851959, 0.694057480199262, 0.550546196289361, 0.525217157322913, 0.212139142677188, 0.862300432287157, 0.465007666498423, 0.390217630658299, 0.250784393632784, 0.728839098243043, 0.757850595051423, 0.985602638451383, 0.129193379776552, 0.119346820283681, 0.247874087886885, 0.0578116811811924, 0.614560918882489, 0.263679290190339, 0.831840219441801, 0.702077423455194, 0.539761470630765, 0.183106842916459, 0.0978696108795702, 0.787322730291635, 0.706133443629369, 0.469276048708707, 0.421247723279521, 0.296663819579408, 0.989214273635298, 0.403854221804067, 0.256786758778617, 0.403823110507801, 0.873210709542036, 0.41881359484978, 0.424839850515127, 0.0868914204183966, 0.758887288160622, 0.758761347969994, 0.396099738776684, 0.250490014208481, 0.776843266794458, 0.643614781089127, 0.544213514775038, 0.710586765082553, 0.971439274493605, 0.132571995723993, 0.707431207876652, 0.848932297201827, 0.310505135683343, 0.098466101102531, 0.950989252654836, 0.638605962740257, 0.709316807100549, 0.221998384455219, 0.336600758833811, 0.256129593821242, 0.574492629617453, 0.986828639637679, 0.828525808872655, 0.631352809257805, 0.607395692262799, 0.776586953084916, 0.876757395453751, 0.811868591234088, 0.922144181560725, 0.754510985454544, 0.909150497522205, 0.947506254073232, 0.172236705198884, 0.962013041600585, 0.643610095605254, 0.229900564532727, 0.649056831374764, 0.291978255379945, 0.368203590158373, 0.770259033190086, 0.00540518830530345, 0.272684170166031, 0.446710997726768, 0.487439919263124, 0.316217475105077, 0.964730207342654, 0.693990624276921, 0.561113970819861, 0.482481942977756, 0.334757950855419, 0.121950790286064, 0.599882940528914, 0.758767848834395, 0.531139024998993, 0.0880152389872819, 0.628411572659388, 0.0287272313144058, 0.935260554077104, 0.325096005341038, 0.798818981274962, 0.0854377308860421, 0.731680589728057, 0.315571876475587, 0.460755710257217, 0.243891360936686, 0.734124839073047, 0.67491690069437, 0.11665814043954, 0.799569312948734, 0.327243498293683, 0.0449250871315598, 0.963041766779497, 0.911213494138792, 0.913298186846077, 0.32982022408396, 0.0679582310840487, 0.611368908546865, 0.262150851543993, 0.296863852534443, 0.5036868408788, 0.397212129086256, 0.70929285348393, 0.0476620541885495, 0.556819583522156, 0.332252991152927, 0.925186224747449, 0.928820671048015, 0.377476670313627, 0.49050212581642, 0.996423732256517, 0.0148738499265164, 0.568940210156143, 0.71840082318522, 0.802870151819661, 0.389945498667657, 0.163962710415944, 0.218604745808989, 0.408486429369077, 0.866374764125794, 0.983747028978541, 0.508430590154603, 0.987439227988943, 0.388555245473981, 0.239320137538016, 0.632683528587222, 0.863338702823967, 0.0335185695439577, 0.216646834975109, 0.0363891627639532, 0.913077889243141, 0.916055401554331, 0.125999386422336, 0.313383843516931, 0.831000489415601, 0.693107610568404, 0.412557960487902, 0.889642220223323, 0.516259075840935, 0.936415531206876, 0.465929167577997, 0.975388077553362, 0.222828032914549, 0.0944923071656376, 0.950885547557846, 0.316720059840009, 0.083851735573262, 0.603858182439581, 0.293060038704425, 0.921307107433677, 0.953540153801441, 0.25370629876852, 0.183968746336177, 0.648116105701774, 0.245553358923644, 0.224571132101119, 0.788999767974019, 0.738069721963257, 0.831053547793999, 0.962683452060446, 0.363975484622642, 0.907984102610499, 0.0331092656124383, 0.392468781443313, 0.642573413439095, 0.616119746817276, 0.53834278928116, 0.34514132537879, 0.859984875191003, 0.322607912356034, 0.183157832361758, 0.584613404236734, 0.313748677726835, 0.580828511156142, 0.378850945737213, 0.798400270985439, 0.537045181263238, 0.904368494870141, 0.923240662086755, 0.758687176508829, 0.292594867525622, 0.795790749369189, 0.565591314341873, 0.277593237580732, 0.999645837116987, 0.260747140040621, 0.800832533510402, 0.648727104300633, 0.228405517758802, 0.203113685594872, 0.670267634792253, 0.720617830753326, 0.868913355050609, 0.147144824732095, 0.787570846267045, 0.282625265186653, 0.975086729275063, 0.62101922580041, 0.585791551275179, 0.0956100176554173, 0.902713528135791, 0.1875480490271, 0.623025460867211, 0.125090706394985, 0.637860344490036, 0.213725840672851, 0.754935199394822, 0.837880351115018, 0.85048757167533, 0.108609510818496, 0.267750641796738, 0.208609394961968, 0.964278666069731, 0.957129742950201, 0.844656610395759, 0.210548165254295, 0.707367569208145, 0.95568578899838, 0.152424469124526, 0.54146617394872, 0.12265466619283, 0.752180464100093, 0.938305499730632, 0.0421730212401599, 0.363666086224839, 0.183457834646106, 0.318108756095171, 0.854390414198861, 0.814429061021656, 0.295351371867582, 0.149973277235404, 0.370767917251214, 0.58153832773678, 0.298797032097355, 0.298293814063072, 0.126896064030007, 0.311406139284372, 0.500161870848387, 0.464918679790571, 0.169859444955364, 0.14185396861285, 0.804442307446152, 0.398634822573513, 0.576428654370829, 0.563581977970898, 0.167862954316661, 0.886601140256971, 0.701393519761041, 0.0430170956533402, 0.610138951567933, 0.872018803609535, 0.010248446604237, 0.95742681925185, 0.996920922305435, 0.628859280608594, 0.499179987469688, 0.494144607335329, 0.816550495335832, 0.536285863257945, 0.237434070790187, 0.614132032264024, 0.155444480944425, 0.0655941148288548, 0.0278662666678429, 0.0245882547460496, 0.832952715922147, 0.601120999548584, 0.613251193892211, 0.846864964812994, 0.602780587039888, 0.845062209526077, 0.718063837848604, 0.137293237727135, 0.759955537039787, 0.825897434027866, 0.304214290343225, 0.219157057348639, 0.768604717450216, 0.386107460362837, 0.134805949870497, 0.973706908291206, 0.520534069510177, 0.556824055267498, 0.385954971658066, 0.767287495313212, 0.912292802939191, 0.826779289869592, 0.512119830586016, 0.369328101165593, 0.20698697376065, 0.708586444612592, 0.370852310443297, 0.743085872847587, 0.211579586844891, 0.625948769273236, 0.102273545926437, 0.168804282788187, 0.0634424122981727, 0.050092113437131, 0.313437074888498, 0.639623458962888, 0.407611269038171, 0.362459101015702, 0.061765540856868, 0.142176856985316, 0.339815753279254, 0.295646052574739, 0.50796634494327, 0.748407789040357, 0.364368901820853, 0.514022697461769, 0.837512924568728, 0.146811286685988, 0.791655919281766, 0.190755368210375, 0.680415719049051, 0.324065638938919, 0.656595850596204, 0.471957703121006, 0.811360173858702, 0.247791892150417, 0.0322266195435077, 0.673520346404985, 0.322668284876272, 0.566043196013197, 0.0427588627208024, 0.814495531376451, 0.660568981431425, 0.138883636100218, 0.124188220128417, 0.84148990851827, 0.416463277768344, 0.709064160240814, 0.689580981153995, 0.758031332399696, 0.219437402207404, 0.553778097033501, 0.975243398919702, 0.602307787863538, 0.376588172279298, 0.557829841272905, 0.0395973487757146, 0.165107402019203, 0.0468014916405082, 0.329389969585463, 0.569013925734907, 0.278481689048931, 0.513453105930239, 0.808262159116566, 0.0305439180228859, 0.581426036776975, 0.656419605948031, 0.45449157548137, 0.417582812951878, 0.700323475524783, 0.174651205772534, 0.137070965021849, 0.878821888007224, 0.511894403956831, 0.760520433774218, 0.502775171538815, 0.962280555162579, 0.340982128400356, 0.88696457981132, 0.706971677020192, 0.915670212823898, 0.359105536015704, 0.713551708264276, 0.725084545789286, 0.169894409133121, 0.987331371987239, 0.0681838125456125, 0.593979073222727, 0.123124807607383, 0.563471011584625, 0.308847267413512, 0.838894790271297, 0.0938488976098597, 0.473303940612823, 0.469092157902196, 0.290592316072434, 0.952549476642162, 0.911423894111067, 0.360453678295016, 0.289149909513071, 0.438028566306457, 0.145675764651969, 0.363204362569377, 0.0782442823983729, 0.0381478269118816, 0.542169166030362, 0.943225571885705, 0.978241698350757, 0.472060150001198, 0.172393365064636, 0.779572107829154, 0.22656657663174, 0.538990733446553, 0.151542813284323, 0.0425935885868967, 0.755974594503641, 0.949834915576503, 0.559561160393059, 0.463874638546258, 0.563632702920586, 0.926799264969304, 0.338715035002679, 0.88241058960557, 0.263209191383794, 0.704809993505478, 0.391704849665985, 0.944030164508149, 0.466350150527433, 0.994423395954072, 0.0967100709676743, 0.62099023303017, 0.806815324816853, 0.219929235521704, 0.642798305256292, 0.545501599321142, 0.105585562996566, 0.566456394502893, 0.420817266684026, 0.132087394362316, 0.50633746618405, 0.744982494041324, 0.00328504992648959, 0.567651536781341, 0.88962886412628, 0.46918768947944, 0.483799990499392, 0.325680330162868, 0.611391612328589, 0.432730111060664, 0.325568552361801, 0.80972487712279, 0.68540523853153, 0.262138119898736, 0.258063618792221, 0.17169145308435, 0.980308268917724, 0.807705818442628, 0.327220435952768, 0.188658529659733, 0.704747125506401, 0.926384467398748, 0.242847903864458, 0.666868816129863, 0.620167418150231, 0.644477068213746, 0.268132655648515, 0.571927722543478, 0.331388340098783, 0.985179882263765, 0.531455902615562, 0.475003345869482, 0.320511323865503, 0.207451376598328, 0.0628073958214372, 0.0238590463995934, 0.481477358145639, 0.681398217799142, 0.183739505708218, 0.730079646687955, 0.281792448367923, 0.0520001908298582, 0.0584355448372662, 0.905496711842716, 0.100314741721377, 0.410075875930488, 0.811082725413144, 0.0516004068776965, 0.105443629901856, 0.986708614509553, 0.487715605413541, 0.844690315192565, 0.0148316039703786, 0.684302859706804, 0.704632092732936, 0.470129266148433, 0.0124706486240029, 0.0361622045747936, 0.508171244291589, 0.0438067647628486, 0.00553501094691455, 0.642123914323747, 0.197533115977421, 0.429208155954257, 0.248741285409778, 0.185135642066598, 0.244540045503527, 0.951639400795102, 0.141848023748025, 0.791768874274567, 0.328146539628506, 0.569266503443941, 0.95178470434621, 0.486081559211016, 0.667171077337116, 0.880210597068071, 0.854843873064965, 0.683581955730915, 0.53859268175438, 0.179902401519939, 0.108441931894049, 0.384173269150779, 0.23578542820178, 0.896518579917029, 0.493431928334758, 0.938353867270052, 0.124518821481615, 0.731457284651697, 0.890182305127382, 0.220730776432902, 0.781682793283835, 0.0461310867685825, 0.545485662994906, 0.531347184441984, 0.716858123894781, 0.764676415361464, 0.70508581539616, 0.636108665261418, 0.20420166477561, 0.723865767009556, 0.682787192752585, 0.0881380680948496, 0.707720032194629, 0.705956674413756, 0.580999742494896, 0.303174235625193, 0.801078439923003, 0.199740113690495, 0.467802806058899, 0.640642159152776, 0.540130688808858, 0.470848120981827, 0.0503276197705418, 0.592244565952569, 0.765154967084527, 0.966655782191083, 0.870691710850224, 0.213516205316409, 0.699006548151374, 0.520738549064845, 0.229878877755255, 0.73696565371938, 0.310767719056457, 0.170352030545473, 0.64393965806812, 0.521434410475194, 0.87786667374894, 0.617553491843864, 0.646503006340936, 0.196563945617527, 0.446843733312562, 0.398842131718993, 0.363591385772452, 0.351878936169669, 0.670794074656442, 0.606174752116203, 0.367966244230047, 0.451167004182935, 0.197526541072875, 0.808539069723338, 0.239489910425618, 0.706188917858526, 0.0889873029664159, 0.848826017929241, 0.238179822918028, 0.454300468089059, 0.576997508993372, 0.22741882246919, 0.0433777500875294, 0.744661716511473, 0.461444343440235, 0.643229339038953, 0.755441719200462, 0.123176392167807, 0.499065053416416, 0.632653938839212, 0.586481043370441, 0.709105486981571, 0.0752072788309306, 0.447426895145327, 0.359966587508097, 0.441380224656314, 0.619165785610676, 0.953048818744719, 0.245582578703761, 0.141329050762579, 0.713749802671373, 0.889789277222008, 0.172117152484134, 0.904754891525954, 0.689850537106395, 0.664635592140257, 0.405116180656478, 0.814118602313101, 0.996108708903193, 0.437554355245084, 0.863736102124676, 0.413650022819638, 0.889935029437765, 0.749339086702093, 0.721607282524928, 0.342189041199163, 0.374767917208374, 0.311500357231125, 0.498640296282247, 0.8364290620666, 0.172311814967543, 0.372296748682857, 0.8549751911778, 0.778524820692837, 0.676700255367905, 0.922284153522924, 0.806165278423578, 0.64199617295526, 0.78643131069839, 0.822548350086436, 0.57421525917016, 0.486072453204542, 0.11414137063548, 0.199891134863719, 0.812219768296927, 0.0436038819607347, 0.645659008529037, 0.714992703637108, 0.641917390050367, 0.897087708115578, 0.139756658580154, 0.132835325086489, 0.0857355406042188, 0.999277296243235, 0.239594074431807, 0.0149586689658463, 0.657411806983873, 0.562801310559735, 0.167653806507587, 0.974274610169232, 0.258520324481651, 0.0992080317810178, 0.961673145182431, 0.561807554680854, 0.794845058582723, 0.188158329110593, 0.691726703196764, 0.864409689325839, 0.470974911469966, 0.368653228972107, 0.68270954512991, 0.871288299793378, 0.679715428268537, 0.851408824324608, 0.858536610612646, 0.488473553210497, 0.680252001387998, 0.151602274971083, 0.9280517953448, 0.989902239060029, 0.430612423224375, 0.635358075844124, 0.0283259872812778, 0.241181954275817, 0.596415231702849, 0.641558651579544, 0.0612392870243639, 0.049524859059602, 0.239962170366198, 0.869284148095176, 0.685266283107921, 0.0962767219170928, 0.757184816291556, 0.824188263621181, 0.872968266485259, 0.709169669775292, 0.341017298633233), actual = c(1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"