Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.6963854        0.7251055 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4886048        0.4886048        0.4931510 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.417331788223237, 0.916102376068011, 0.296034869737923, 0.752982285805047, 0.155759556451812, 0.270794182317331, 0.613834351301193, 0.837461778894067, 0.718040631152689, 0.0824114228598773, 0.0261160060763359, 0.77419492113404, 0.753449971787632, 0.431437860010192, 0.298610958037898, 0.129015712765977, 0.700626592617482, 0.251864892430604, 0.203511729370803, 0.560998945264146, 0.672785990638658, 0.426553559722379, 0.160560336662456, 0.783666837029159, 0.407445141347125, 0.525491680018604, 0.960492981597781, 0.333699288778007, 0.594102546107024, 0.903954399051145, 0.533723691711202, 0.646861901273951, 0.83628723048605, 0.332723877858371, 0.549962241435423, 0.521431452361867, 0.19060414400883, 0.323363526258618, 0.78916848776862, 0.347697584656999, 0.164895326597616, 0.250750969396904, 0.727364465361461, 0.304938661633059, 0.687113082967699, 0.167384335072711, 0.280806274386123, 0.310974556719884, 0.2186077341903, 0.849510099738836, 0.171795894159004, 0.192062764894217, 0.0178804837632924, 0.993796211667359, 0.565333608305082, 0.821589353028685, 0.191341607132927, 0.0597667030524462, 0.552831118227914, 0.296701188199222, 0.306921398034319, 0.621505500981584, 0.240576571086422, 0.146564261987805, 0.400991619564593, 0.160643008537591, 0.0890872827731073, 0.10542298364453, 0.118579154834151, 0.513128134422004, 0.790753426495939, 0.0304109507706016, 0.448905253084376, 0.480086180614308, 0.996061127865687, 0.451654787175357, 0.353062083711848, 0.827589143998921, 0.903823970584199, 0.128300483338535, 0.560143543407321, 0.590932219522074, 0.0486166086047888, 0.878998465370387, 0.291623454773799, 0.498870007926598, 0.738281976198778, 0.219796794699505, 0.914880070602521, 0.623326746281236, 0.73605494434014, 0.698232715949416, 0.971230083610862, 0.0140032530762255, 0.915966283995658, 0.278543523512781, 0.0732282586395741, 0.646860865643248, 0.459183721337467, 0.0938384556211531, 0.797579369274899, 0.620203790720552, 0.994971962878481, 0.137040649773553, 0.767774926731363, 0.23861145763658, 0.920883002923802, 0.505088709061965, 0.422925936290994, 0.728536549024284, 0.951712252572179, 0.428159324917942, 0.993161020800471, 0.706295871408656, 0.278784220572561, 0.236807205481455, 0.893585399026051, 0.602308840723708, 0.484232593793422, 0.0832717344164848, 0.924446911318228, 0.752381632104516, 0.870030449703336, 0.320764807518572, 0.491588764125481, 0.299116056412458, 0.226022662129253, 0.744206477189437, 0.241405638633296, 0.157647872343659, 0.70566641795449, 0.279155462514609, 0.3689490549732, 0.232064741896465, 0.674669282510877, 0.221544002415612, 0.436342688044533, 0.604589117923751, 0.301499193301424, 0.951831158949062, 0.679572265828028, 0.885718709789217, 0.685987691162154, 0.214550813660026, 0.8777809177991, 0.0155116443056613, 0.615342975826934, 0.918475690064952, 0.966342157684267, 0.980368397431448, 0.00491599179804325, 0.436512728920206, 0.706574402749538, 0.445798004744574, 0.691521897213534, 0.738619682379067, 0.0349455554969609, 0.670882477657869, 0.523637441452593, 0.436522626550868, 0.467350855702534, 0.995548677630723, 0.0699521587230265, 0.821388103766367, 0.63206102931872, 0.321251280372962, 0.144126612693071, 0.251142050139606, 0.475922290235758, 0.198019115952775, 0.4041334874928, 0.535717660794035, 0.872354943072423, 0.455512905260548, 0.558139371452853, 0.0259994072839618, 0.952205228619277, 0.69700898299925, 0.0300859913695604, 0.817757990444079, 0.291278164368123, 0.277929757256061, 0.313464512117207, 0.914716649800539, 0.244323036866263, 0.569825583603233, 0.518082862487063, 0.150522635085508, 0.765799468383193, 0.449564890470356, 0.615555653581396, 0.914074159692973, 0.419216064270586, 0.464134208392352, 0.304826283594593, 0.699924159329385, 0.664703999180347, 0.570063398685306, 0.513817427912727, 0.234403110574931, 0.077227332862094, 0.947911031544209, 0.509447154123336, 0.785815523704514, 0.427134096855298, 0.446227630134672, 0.919205725891516, 0.77424835646525, 0.482887219637632, 0.0822028538677841, 0.0514624479692429, 0.981454926775768, 0.0317796522285789, 0.112383509287611, 0.462231585057452, 0.170386733720079, 0.58309034164995, 0.528411238221452, 0.642751554958522, 0.597977447789162, 0.254253220744431, 0.0729353853967041, 0.848900166573003, 0.401789560448378, 0.813827898120508, 0.276524234097451, 0.842087991070002, 0.345108441542834, 0.687264606356621, 0.238897516857833, 0.764055476756766, 0.100673884851858, 0.4778538858518, 0.61001356318593, 0.731250846292824, 0.278786294860765, 0.162342906231061, 0.0286991223692894, 0.695934485178441, 0.761955960420892, 0.629907942377031, 0.46609994978644, 0.269886772381142, 0.970696491654962, 0.0360179943963885, 0.931312136352062, 0.724237965652719, 0.546088226838037, 0.901425767689943, 0.751798133132979, 0.502114181406796, 0.441444738768041, 0.886472437065095, 0.878132423851639, 0.297403384931386, 0.915466579608619, 0.340149712515995, 0.403381591895595, 0.255783412605524, 0.341568962438032, 0.200503444531932, 0.490268850931898, 0.0180651792325079, 0.236761942040175, 0.558870530221611, 0.662656377535313, 0.599356378894299, 0.178376383380964, 0.673575797816738, 0.212084547616541, 0.263247593306005, 0.859581452794373, 0.425062003778294, 0.876854313304648, 0.863427378702909, 0.957458705175668, 0.91280809068121, 0.197343214647844, 0.20136025804095, 0.935582301579416, 0.380785066401586, 0.95321078854613, 0.437248381553218, 0.403053886722773, 0.917285584378988, 0.951139128301293, 0.305585918715224, 0.0382491119671613, 0.225394737208262, 0.125518034677953, 0.951918094884604, 0.791164927883074, 0.911362902261317, 0.276710780104622, 0.11027211160399, 0.103840752504766, 0.58867886569351, 0.569361778674647, 0.674312934745103, 0.111312560504302, 0.273548342753202, 0.397662111092359, 0.66732945269905, 0.835901292273775, 0.366639716783538, 0.230714912526309, 0.841376035939902, 0.890447662677616, 0.801312421681359, 0.341910931514576, 0.679958950728178, 0.889833164168522, 0.862033111508936, 0.828206845326349, 0.425128423376009, 0.247794605558738, 0.478083667578176, 0.354019543388858, 0.0257113566622138, 0.703266781987622, 0.119646067498252, 0.241063865367323, 0.940966598456725, 0.656321126734838, 0.940368376672268, 0.200742477318272, 0.776480143889785, 0.8753606765531, 0.594030450098217, 0.114999663084745, 0.895374858984724, 0.146986357402056, 0.93644956452772, 0.122113701654598, 0.128684030147269, 0.633213800610974, 0.437596798874438, 0.00640821177512407, 0.156873210100457, 0.505454403348267, 0.515142503194511, 0.656226235209033, 0.567236168542877, 0.0123636461794376, 0.265180011279881, 0.525589981582016, 0.33092405507341, 0.829456241102889, 0.204944679047912, 0.505703567760065, 0.254121768055484, 0.459850418381393, 0.0844407263211906, 0.250072412891313, 0.522542702034116, 0.469923339318484, 0.667262028437108, 0.221853700932115, 0.337640271522105, 0.516786873107776, 0.219711065059528, 0.221682968083769, 0.479746408993378, 0.544921547407284, 0.59250169317238, 0.909786133794114, 0.538000602042302, 0.00880705728195608, 0.204257164616138, 0.316817077342421, 0.546555619919673, 0.692156377481297, 0.28482706611976, 0.576396559830755, 0.0496563648339361, 0.973530151881278, 0.331869866931811, 0.9563197940588, 0.862968601286411, 0.778451906517148, 0.387502447934821, 0.286226395983249, 0.130708752898499, 0.0152413831092417, 0.471593576716259, 0.872563397977501, 0.802499170415103, 0.29426185740158, 0.697626527631655, 0.119267075555399, 0.915534871164709, 0.122719368198887, 0.588823798811063, 0.17919860035181, 0.0191193050704896, 0.747228110674769, 0.171793832909316, 0.9312333392445, 0.323430591495708, 0.218335861107334, 0.0164398329798132, 0.522362247342244, 0.082799335476011, 0.715110026067123, 0.930085508851334, 0.403298140037805, 0.246375460876152, 0.17615287960507, 0.653482155641541, 0.424746132921427, 0.602660472039133, 0.415137795731425, 0.76239821780473, 0.0533918298315257, 0.0686689862050116, 0.976184032158926, 0.429004368139431, 0.326102938270196, 0.358509728452191, 0.590233528520912, 0.841480401344597, 0.0986943326424807, 0.722947753267363, 0.713371637510136, 0.962112339446321, 0.00659299036487937, 0.498584051849321, 0.165120498510078, 0.0012135140132159, 0.970733921276405, 0.11657879082486, 0.867167680989951, 0.565818583825603, 0.332915783394128, 0.277083321474493, 0.416437108302489, 0.438281283946708, 0.350069473031908, 0.137195045128465, 0.433534317882732, 0.745154168922454, 0.734562035417184, 0.178322444902733, 0.350460068555549, 0.218147421255708, 0.12449409160763, 0.873530178098008, 0.642236947547644, 0.374191784998402, 0.867915772134438, 0.955655119149014, 0.60308035206981, 0.15002597658895, 0.478262932039797, 0.198370232479647, 0.998449889244512, 0.0914447910618037, 0.977824796922505, 0.054839264601469, 0.140996590489522, 0.585133995860815, 0.477064729901031, 0.548680342966691, 0.0727038094773889, 0.480938971508294, 0.407983350334689, 0.60990216373466, 0.230178026948124, 0.170526012778282, 0.848850642563775, 0.135542654199526, 0.378100306726992, 0.890980710042641, 0.354869070928544, 0.882109037600458, 0.028781182365492, 0.897601218661293, 0.103023640112951, 0.702832160983235, 0.360428464133292, 0.495181452017277, 0.358514487044886, 0.317756256787106, 0.270422189729288, 0.655606344342232, 0.35112539306283, 0.999626255594194, 0.936771000968292, 0.302876964677125, 0.710946759674698, 0.948017020942643, 0.132964081363752, 0.892279397463426, 0.995875865919515, 0.16808448638767, 0.0458765202201903, 0.237230800557882, 0.675745686283335, 0.823533731512725, 0.718717258190736, 0.8610037737526, 0.874325396027416, 0.65854211547412, 0.382996208034456, 0.47069977177307, 0.131649622693658, 0.600864527514204, 0.0879188128747046, 0.448183862492442, 0.724395781056955, 0.421993682859465, 0.286181164439768, 0.0795184762682766, 0.86106606759131, 0.507589080603793, 0.417042914079502, 0.406630084384233, 0.166816007578745, 0.677715366706252, 0.353264887118712, 0.0322948715183884, 0.161659821867943, 0.363447046838701, 0.944279461633414, 0.0826102185528725, 0.61557339457795, 0.901093003805727, 0.463568523526192, 0.175259017618373, 0.279483787482604, 0.14317942946218, 0.15879153739661, 0.350163068156689, 0.19205662002787, 0.24170665605925, 0.0687070626299828, 0.274311516433954, 0.345047240843996, 0.306287825107574, 0.072091757087037, 0.104859301820397, 0.525165402330458, 0.583488282514736, 0.830945623805746, 0.588460898259655, 0.322069956921041, 0.838666363852099, 0.488845797022805, 0.255666061071679, 0.897736291168258, 0.463812410598621, 0.210034880088642, 0.524874444352463, 0.57136775739491, 0.475809658644721, 0.94359744922258, 0.00813414715230465, 0.108831197256222, 0.384271716699004, 0.896780072245747, 0.491588761098683, 0.283123358385637, 0.467710670549423, 0.601649799151346, 0.740137842716649, 0.0352043649181724, 0.792186892824247, 0.561817587353289, 0.185957228764892, 0.33059615874663, 0.361538539174944, 0.382494677789509, 0.470911832991987, 0.498695211252198, 0.915176383918151, 0.158015406690538, 0.0981824048794806, 0.745694148354232, 0.730208162683994, 0.916110596386716, 0.452633470762521, 0.758590148994699, 0.695621023420244, 0.741047564893961, 0.28483567154035, 0.887482194695622, 0.110260971589014, 0.656012266408652, 0.282220302615315, 0.609724723733962, 0.00870941951870918, 0.392230351222679, 0.261488799937069, 0.814689758932218, 0.696033469634131, 0.381504581775516, 0.960707580437884, 0.349718089448288, 0.442711926298216, 0.131016989704221, 0.459342152345926, 0.283413502853364, 0.289359341841191, 0.935884896665812, 0.272939933463931, 0.303405968472362, 0.245939088985324, 0.476367973955348, 0.10800842801109, 0.232268295716494, 0.0850597159005702, 0.98195837601088, 0.620570091763511, 0.0217869421467185, 0.329297793796286, 0.639543124707416, 0.52648917445913, 0.813951030606404, 0.398969451664016, 0.670542157720774, 0.178375905845314, 0.0755261965095997, 0.162071745144203, 0.780273505486548, 0.361878033727407, 0.530499007320032, 0.880913441302255, 0.381779800402001, 0.7762543160934, 0.992517825681716, 0.807322109118104, 0.800117868930101, 0.144949107896537, 0.467644728487357, 0.321313234977424, 0.492604803293943, 0.196321798954159, 0.213709150440991, 0.461088297422975, 0.262134145013988, 0.267188989790156, 0.978585612494498, 0.190116020152345, 0.624298456124961, 0.573392940219492, 0.802016142988577, 0.350323259364814, 0.771899212850258, 0.965265801874921, 0.866424039704725, 0.817932611331344, 0.106422543758526, 0.421300523448735, 0.99535553320311, 0.908281381474808, 0.778809421462938, 0.826263841008767, 0.849566122051328, 0.411676167277619, 0.38620293000713, 0.56534350826405, 0.645458213752136, 0.333762690192088, 0.845161560922861, 0.175797735340893, 0.716328586917371, 0.383343606255949, 0.426234438549727, 0.00852909265086055, 0.196885640267283, 0.573578208684921, 0.420775479404256, 0.965596193447709, 0.696627025958151, 0.0073981168679893, 0.969216174911708, 0.347136378986761, 0.267071335576475, 0.535190584836528, 0.435200414853171, 0.367055653128773, 0.904761691112071, 0.57338760048151, 0.189367761602625, 0.889089034870267, 0.251339769223705, 0.141929495846853, 0.584781061857939, 0.204240994062275, 0.566495859529823, 0.854967347811908, 0.850616571027786, 0.70321947010234, 0.476903266040608, 0.811378618469462, 0.774002853548154, 0.95739247277379, 0.801432776032016, 0.206886756699532, 0.0270909392274916, 0.329447924159467, 0.959541401127353, 0.790497880429029, 0.851137814344838, 0.786581295309588, 0.124072170350701, 0.71295986417681, 0.910701057640836, 0.0301770374644548, 0.0862859315238893, 0.732953472994268, 0.391401496017352, 0.876247256062925, 0.487964376341552, 0.167590525466949, 0.111678342102095, 0.68409622553736, 0.361761318519711, 0.973997148917988, 0.460431872867048, 0.697728639701381, 0.940784544683993, 0.339328935137019, 0.416838735342026, 0.380675859283656, 0.992450441699475, 0.169851401355118, 0.542773026973009, 0.347886169329286, 0.805158662376925, 0.155055371345952, 0.563746204832569, 0.669035675469786, 0.821107221534476, 0.195208861026913, 0.0267544582020491, 0.486316702328622, 0.377063416410238, 0.350337702548131, 0.293879586737603, 0.645947150886059, 0.644456934183836, 0.561816534493119, 0.754952520364895, 0.867995209759101, 0.714762555900961, 0.634360040305182, 0.257679190021008, 0.674509732285514, 0.47205925360322, 0.384763628011569, 0.306063179159537, 0.932634885655716, 0.576945251086727, 0.374347362434492, 0.52118536666967, 0.40201832447201, 0.111077962210402, 0.972023510606959, 0.569025550270453, 0.115179031621665, 0.165759459836408, 0.175253280671313, 0.924775128718466, 0.318053754046559, 0.984412071527913, 0.182818900095299, 0.284408238949254, 0.938883543480188, 0.0216107510495931, 0.364260280504823, 0.368050490738824, 0.0621394296176732, 0.407013306394219, 0.315565491328016, 0.195580960717052, 0.942849248182029, 0.58995466819033, 0.980039975605905, 0.255700312787667, 0.286442124517635, 0.939218459418043, 0.662385492352769, 0.154183007311076, 0.721684025367722, 0.352238285588101, 0.907807088922709, 0.117406793870032, 0.0349110125098377, 0.131858235225081, 0.70641771890223, 0.218933037249371, 0.288009729236364, 0.607286346377805, 0.520677205640823, 0.982059977948666, 0.117057852214202, 0.854451154824346, 0.521577165462077, 0.665504658827558, 0.414578275987878, 0.230850259540603, 0.611921756062657, 0.457564588403329, 0.965408748015761, 0.148758828872815, 0.306388306431472, 0.464153153588995, 0.26464714971371, 0.146579733118415, 0.796073525911197, 0.510356053942814, 0.896378571400419, 0.28668341669254, 0.328502402175218, 0.20238416432403, 0.619897317374125, 0.634809852810577, 0.761610268615186, 0.103056683205068, 0.609243023907766, 0.107752203475684, 0.983973002526909, 0.0551311238668859, 0.592496986268088, 0.841796244028956, 0.573104537790641, 0.833224676316604, 0.166069189086556, 0.0971165127120912, 0.231482960050926, 0.866777844028547, 0.765557112172246, 0.468387476634234, 0.891913265455514, 0.409634046722203, 0.758073071716353, 0.976196730742231, 0.536220802227035, 0.661198867717758, 0.337783937109634, 0.208977398695424, 0.0344001771882176, 0.511517848586664, 0.977418116992339, 0.448156855301932, 0.45833080378361, 0.413365062326193, 0.596435371786356, 0.804070037556812, 0.651681960793212, 0.208786136237904, 0.206453016260639, 0.845835763728246, 0.56291104783304, 0.196157327620313, 0.612922886619344, 0.491689376765862, 0.233840616419911, 0.17207113141194, 0.126445874106139, 0.649882694939151, 0.518145764712244, 0.939185566734523, 0.972844145493582, 0.238893853966147, 0.463484486332163, 0.404544050106779, 0.0975651054177433, 0.185738625004888, 0.254153062589467, 0.00821558805182576, 0.563101237406954, 0.158498489065096, 0.779374996665865, 0.413102371385321, 0.811206053476781, 0.236202894477174, 0.561787232290953, 0.367788636591285, 0.112305516609922, 0.397491029230878, 0.9575511936564, 0.239005260867998, 0.832983464468271, 0.220814459957182, 0.726738350931555, 0.113135135499761, 0.762024787720293, 0.272924544056877, 0.513162920949981, 0.453339830040932, 0.283959742868319, 0.189479061868042, 0.603002277435735, 0.942909591132775, 0.3865689530503, 0.839312298921868, 0.9826211030595, 0.496833343990147, 0.439636850496754, 0.00871116225607693, 0.15418834076263, 0.411648745881394, 0.492981330258772, 0.509091520449147, 0.63847952731885, 0.634457705076784, 0.0976798289921135, 0.230094737838954, 0.928734170040116, 0.829345709877089, 0.0768389841541648, 0.987476505571976, 0.132245438871905, 0.385472386376932, 0.996415596920997, 0.0123492702841759, 0.0137743463274091, 0.0187023973558098, 0.137283333344385, 0.640466823242605, 0.273559014778584, 0.239315680461004, 0.425052038626745, 0.380580089287832, 0.254955774871632, 0.544256601948291, 0.487200365168974, 0.459458877099678, 0.569578970549628, 0.569409134332091, 0.613470753189176, 0.215314943576232, 0.234342637471855, 0.973809607094154, 0.846947662299499, 0.87315822346136, 0.730933761224151, 0.458200634457171, 0.545281550148502, 0.988985070260242, 0.175399879924953, 0.434240506729111, 0.170868766261265, 0.341198829002678, 0.538054690463468, 0.192181279184297, 0.806698340456933, 0.107698908308521, 0.743353005265817, 0.370645369868726, 0.0332348039373755, 0.251736310776323, 0.817937936168164, 0.0248844036832452, 0.505244480445981, 0.235878554172814, 0.879020459018648, 0.985598550178111, 0.757680550217628, 0.228099825792015, 0.518759200582281, 0.27200435497798, 0.364271223545074, 0.0236534522846341, 0.20109334285371, 0.840402374044061, 0.204027744708583, 0.932895133271813, 0.700297921895981, 0.497537184739485, 0.19857438467443, 0.942628406453878, 0.735980680212379, 0.122071822639555, 0.791147438110784, 0.483084266539663, 0.178724845871329, 0.597619246458635, 0.907292620744556, 0.906736586708575, 0.669846954289824, 0.648045492824167, 0.477688991231844, 0.574773392174393, 0.359060506802052, 0.961596440291032, 0.97428926313296, 0.974767573177814, 0.899752415250987, 0.610891417367384, 0.812685701996088, 0.104993597371504, 0.0266388866584748, 0.237135037546977, 0.394293045857921, 0.649975313106552, 0.253025792306289, 0.775535664521158), actual = c(1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.995543023338541, 0.434501833748072, 0.819566205143929, 0.88716551172547, 0.610434378497303, 0.765955187613145, 0.89863766124472, 0.497515770606697, 0.988380464259535, 0.570654324023053, 0.808170137926936, 0.188713921466842, 0.961513712303713, 0.299005345907062, 0.891553451307118, 0.83448623190634, 0.796291461680084, 0.543845033738762, 0.791514049749821, 0.829973477637395, 0.816073488211259, 0.976760161109269, 0.0990110207349062, 0.383275442058221, 0.889671068172902, 0.605568591272458, 0.857537110103294, 0.941493784543127, 0.655879223253578, 0.186557905981317, 0.613534491276368, 0.0218923343345523, 0.216785497730598, 0.933193122269586, 0.63901935191825, 0.110704181017354, 0.371027200017124, 0.881043548462912, 0.774799451464787, 0.102359239477664, 0.165236679837108, 0.892436432885006, 0.556354736210778, 0.00513603235594928, 0.965572920395061, 0.115576716605574, 0.616451770532876, 0.570803080918267, 0.0509736172389239, 0.302890898659825, 0.653019251069054, 0.576841784408316, 0.263225336559117, 0.962599145947024, 0.881439546821639, 0.965376703767106, 0.643941241083667, 0.0677843634039164, 0.33949328889139, 0.49649238307029, 0.036211802624166, 0.622448552865535, 0.0167558162938803, 0.653069771593437, 0.83594427537173, 0.019231358775869, 0.531261834083125, 0.215626961318776, 0.336119549814612, 0.198296092217788, 0.266632210230455, 0.351282259915024, 0.0973225717898458, 0.77022890932858, 0.942340224748477, 0.562780449399725, 0.702695638174191, 0.1551697361283, 0.76379434671253, 0.949945857049897, 0.855047623626888, 0.459667245391756, 0.7845705577638, 0.45622984948568, 0.429770896444097, 0.830961973639205, 0.590797711163759, 0.3014205112122, 0.0133954950142652, 0.442171318456531, 0.12624594499357, 0.514263855526224, 0.0148534174077213, 0.263029927853495, 0.919585953466594, 0.727249255403876, 0.786409806227311, 0.859776637051255, 0.605976153863594, 0.654695093864575, 0.445643074344844, 0.915840581292287, 0.0553871493320912, 0.878933063242584, 0.107787524117157, 0.740630388492718, 0.0349836908280849, 0.476095056161284, 0.979764482239261, 0.107807355700061, 0.424901005113497, 0.229930992936715, 0.44951805844903, 0.132352800341323, 0.33749103709124, 0.757113590603694, 0.859377101296559, 0.463984899688512, 0.46382099436596, 0.971102019306272, 0.50245091272518, 0.325545409228653, 0.351806187769398, 0.276459872024134, 0.370249543106183, 0.723402472678572, 0.490755826001987, 0.708727997494861, 0.281230432214215, 0.893317208392546, 0.919506861129776, 0.377269986551255, 0.0992522942833602, 0.741559609770775, 0.581590136513114, 0.137681987136602, 0.482357128057629, 0.183069682214409, 0.132559946272522, 0.159307536436245, 0.60366868134588, 0.353600443340838, 0.873221812304109, 0.941519081359729, 0.0902244695462286, 0.124465425033122, 0.279388369061053, 0.374115670099854, 0.0614355925936252, 0.109986070543528, 0.0790136272553355, 0.952198887942359, 0.246845826040953, 0.760940689826384, 0.726652899757028, 0.0818505131173879, 0.745703799882904, 0.822880893014371, 0.449152847519144, 0.459054538980126, 0.753039738629013, 0.5119082655292, 0.460384895093739, 0.225711394101381, 0.175256109563634, 0.472953679505736, 0.226053112652153, 0.100484796799719, 0.495634161867201, 0.214200595626608, 0.0963806486688554, 0.0796100662555546, 0.0241204956546426, 0.274605685379356, 0.53572592837736, 0.186288569588214, 0.566162323346362, 0.448741475353017, 0.717742886394262, 0.675711981486529, 0.104937203926966, 0.145013784524053, 0.0568513358011842, 0.225028054555878, 0.0389139680191875, 0.127749490085989, 0.66074006841518, 0.216350186383352, 0.64087426266633, 0.158124673180282, 0.12699325545691, 0.263533900491893, 0.675371842924505, 0.0903778236825019, 0.369635911658406, 0.286209793062881, 0.790946622611955, 0.796091500669718, 0.0819762668106705, 0.725718831177801, 0.202066135592759, 0.82941828854382, 0.533058658707887, 0.301925973966718, 0.299701628508046, 0.0632862488273531, 0.60800030012615, 0.651538873789832, 0.912986977491528, 0.298666573362425, 0.995829788735136, 0.271558952983469, 0.425405589397997, 0.874104450922459, 0.98830089555122, 0.216317584272474, 0.975848914124072, 0.885541379917413, 0.709502219455317, 0.366712245158851, 0.525939539540559, 0.535608075093478, 0.52363019878976, 0.188860326539725, 0.255701150512323, 0.731779102003202, 0.644151064800099, 0.75392653234303, 0.269229585537687, 0.0990328122861683, 0.582709626527503, 0.334441638085991, 0.45880227140151, 0.629874135600403, 0.305132644018158, 0.144991419278085, 0.113614117028192, 0.730120491236448, 0.380571174435318, 0.333144186297432, 0.20870515354909, 0.277438449906185, 0.805877136066556, 0.504279892658815, 0.642668206244707, 0.470273471903056, 0.106565200258046, 0.276917273644358, 0.892738596070558, 0.287461389321834, 0.715470475144684, 0.28159209783189, 0.671695317607373, 0.729953353991732, 0.961216791532934, 0.246016350807622, 0.619524866808206, 0.350416224449873, 0.125882215797901, 0.60053987801075, 0.8019182048738, 0.602840627077967, 0.252412896603346, 0.00992193864658475, 0.460904944920912, 0.703844045288861, 0.588737979298458, 0.401803531916812, 0.526915666414425, 0.961820802418515, 0.575079978676513, 0.67530326731503, 0.72817989019677, 0.552601563045755, 0.218316601356491, 0.478897424880415, 0.320102421566844, 0.486332176718861, 0.443411767249927, 0.987437221920118, 0.853228074498475, 0.719437581719831, 0.649324203142896, 0.0619108825922012, 0.655685519333929, 0.266455289442092, 0.849941654130816, 0.826685706386343, 0.712381685385481, 0.509939573006704, 0.0603453840594739, 0.802631936501712, 0.974446725798771, 0.637698894366622, 0.158418184146285, 0.864143531536683, 0.269928817404434, 0.207706131972373, 0.802440620725974, 0.134510764619336, 0.0853862850926816, 0.944733090000227, 0.383190031861886, 0.703866286901757, 0.776023829588667, 0.290056412806734, 0.161079949233681, 0.926717159803957, 0.378413381753489, 0.0195320702623576, 0.715475185541436, 0.756366381654516, 0.421031365636736, 0.0536279701627791, 0.704505500849336, 0.198232057970017, 0.430024345172569, 0.69845197419636, 0.584793797694147, 0.791407273849472, 0.0668416738044471, 0.572280112886801, 0.0779901430942118, 0.13045019027777, 0.750316657591611, 0.283082572743297, 0.17577898548916, 0.897888759151101, 0.413381641497836, 0.729972108034417, 0.61772444145754, 0.772084933239967, 0.0577740322332829, 0.584202158730477, 0.641571998363361, 0.535954171326011, 0.650582193862647, 0.109339293092489, 0.208740539383143, 0.995010879822075, 0.174148411257192, 0.650710885878652, 0.539485883433372, 0.499302584677935, 0.459950571646914, 0.501373340375721, 0.147701016627252, 0.588823473313823, 0.099109003553167, 0.609832091722637, 0.427390116965398, 0.495405537541956, 0.945670247310773, 0.306754906196147, 0.486468398012221, 0.901808527531102, 0.794326388742775, 0.13400374725461, 0.0293305937666446, 0.960219091037288, 0.362217887770385, 0.584279891336337, 0.142242459580302, 0.158764080144465, 0.025614159880206, 0.895109745441005, 0.165159873198718, 0.91290476359427, 0.764671525452286, 0.701319631189108, 0.338724257657304, 0.819355722982436, 0.353877599351108, 0.532460233662277, 0.177881662733853, 0.0244862029794604, 0.558812226634473, 0.104605151107535, 0.599414172582328, 0.112417869269848, 0.571535246679559, 0.142476623179391, 0.688854895997792, 0.970771596999839, 0.725344198988751, 0.354224918875843, 0.497381525114179, 0.973055768292397, 0.891669576754794, 0.85299784434028, 0.859345472184941, 0.795498714549467, 0.106823339825496, 0.620855487184599, 0.169531693216413, 0.222453702706844, 0.627942276652902, 0.693323460407555, 0.81451780628413, 0.111005315789953, 0.138655890244991, 0.832801356213167, 0.303327528294176, 0.919340908993036, 0.788022429449484, 0.732276428490877, 0.794382431544363, 0.871581224957481, 0.0728115756064653, 0.492558081634343, 0.861892016604543, 0.380452115554363, 0.96491818362847, 0.603576983325183, 0.226823881966993, 0.0699626153800637, 0.42095140251331, 0.306705564027652, 0.379875783110037, 0.351095759309828, 0.29498466802761, 0.965046035125852, 0.999220259487629, 0.958616922842339, 0.377101252321154, 0.463030748534948, 0.558362703770399, 0.805953171569854, 0.987785426899791, 0.125796583481133, 0.0292743223253638, 0.991026860196143, 0.329008967615664, 0.318991217995062, 0.179996869526803, 0.00298577500507236, 0.963965042494237, 0.981658262666315, 0.495586903998628, 0.46644476428628, 0.338922832626849, 0.547383312601596, 0.274636540096253, 0.191450754646212, 0.247746577486396, 0.659219683613628, 0.8107452490367, 0.284963393583894, 0.516235998831689, 0.260908500058576, 0.674984979908913, 0.856587316840887, 0.85148477461189, 0.63705795048736, 0.699019862571731, 0.340582089032978, 0.0643389476463199, 0.378827217500657, 0.670077504590154, 0.904678639722988, 0.136991068487987, 0.0733325034379959, 0.0623776311986148, 0.384821596555412, 0.186709677800536, 0.351088139228523, 0.241999322781339, 0.00579835707321763, 0.672516491729766, 0.530778948683292, 0.640826905844733, 0.906162397935987, 0.24193616467528, 0.323799550766125, 0.155919263139367, 0.349515502806753, 0.427218678407371, 0.491573009639978, 0.185686937300488, 0.92678326764144, 0.550255942391232, 0.359764816937968, 0.91714903479442, 0.464712791377679, 0.544178216950968, 0.0382591052912176, 0.994954289635643, 0.206513216020539, 0.612767083570361, 0.423682923661545, 0.623397898161784, 0.572988829342648, 0.134131891652942, 0.16211234848015, 0.735328427050263, 0.743353570112959, 0.485693895258009, 0.383814131608233, 0.866490927292034, 0.261891578324139, 0.882138836896047, 0.944086622213945, 0.16710270754993, 0.566708120284602, 0.500096667092294, 0.00191050465218723, 0.220101988175884, 0.610450356034562, 0.842717975378036, 0.93788713007234, 0.199490224476904, 0.888962741941214, 0.675679500214756, 0.168892679736018, 0.586509626358747, 0.45121773192659, 0.238961877301335, 0.420237697660923, 0.795935806585476, 0.62207172694616, 0.678708367049694, 0.840572747867554, 0.949052340816706, 0.0453723394311965, 0.541968339355662, 0.953593017766252, 0.0980421486310661, 0.367196469334885, 0.675298724323511, 0.702509315218776, 0.0308677135035396, 0.581917637260631, 0.786664094543085, 0.608066769549623, 0.446492824936286, 0.221564349485561, 0.762562742456794, 0.0963787471409887, 0.570448262151331, 0.456266456982121, 0.360284176189452, 0.967231372138485, 0.796604607487097, 0.863442035857588, 0.450829268665984, 0.364485098980367, 0.961002356605604, 0.418925456702709, 0.325739796739072, 0.00509289419278502, 0.4240904932376, 0.298692802665755, 0.150970801012591, 0.621515234233811, 0.632229371694848, 0.951418608892709, 0.955462850863114, 0.775775183923542, 0.248475478496403, 0.351254401728511, 0.551610282622278, 0.815051961923018, 0.233218572800979, 0.140373511239886, 0.0512777224648744, 0.246902250451967, 0.327946657082066, 0.982062490424141, 0.415684690698981, 0.335173388710245, 0.608179712435231, 0.640317077049986, 0.645448639756069, 0.872475585201755, 0.196678968379274, 0.606869199546054, 0.492919905344024, 0.331806272035465, 0.530201366171241, 0.0598289382178336, 0.285683315945789, 0.202769812196493, 0.913130599306896, 0.941134905209765, 0.922801095992327, 0.168441182700917, 0.285263145808131, 0.239705736283213, 0.112087923102081, 0.404393186094239, 0.271767601137981, 0.329488325398415, 0.737703504506499, 0.719286710489541, 0.382257146062329, 0.870588949648663, 0.965174689423293, 0.430635647848248, 0.65325330151245, 0.52450402197428, 0.528791304677725, 0.524434769758955, 0.667192707769573, 0.672088489867747, 0.327619866468012, 0.594254801981151, 0.131824392592534, 0.727035623043776, 0.222058282466605, 0.55439060786739, 0.394341815030202, 0.90670716855675, 0.884042856981978, 0.917672588955611, 0.131836934946477, 0.992807481205091, 0.170811702497303, 0.464330184739083, 0.592173239449039, 0.389717519516125, 0.245441592065617, 0.267356873955578, 0.018662471557036, 0.623698835493997, 0.849583095870912, 0.280271871713921, 0.510973533149809, 0.354562791762874, 0.862278894754127, 0.513812448596582, 0.496929093496874, 0.493852681014687, 0.158819647505879, 0.16667553409934, 0.911253166152164, 0.639499282231554, 0.419575181556866, 0.792706800391898, 0.558183142682537, 0.0127978336531669, 0.846072318032384, 0.281651497352868, 0.256866040406749, 0.965503240702674, 0.793989053927362, 0.0542709324508905, 0.791666406206787, 0.686127010732889, 0.560020013479516, 0.659565134905279, 0.322877374710515, 0.698200771585107, 0.335423390148208, 0.807328811613843, 0.370759788900614, 0.778775017941371, 0.374762854073197, 0.376260698074475, 0.256799908354878, 0.249345726566389, 0.118607500568032, 0.294888670789078, 0.459791938774288, 0.0652229832485318, 0.0995966556947678, 0.551369710126892, 0.367864507017657, 0.372309022583067, 0.498269321164116, 0.493029663339257, 0.521731766173616, 0.505370083497837, 0.460446861572564, 0.356325291329995, 0.924739946611226, 0.342248692642897, 0.702310691354796, 0.93314636615105, 0.648623807355762, 0.191628215368837, 0.493622443173081, 0.237992986338213, 0.431539627956226, 0.614602839574218, 0.892296144971624, 0.148636183002964, 0.422828255454078, 0.855603080010042, 0.860386245185509, 0.52599586150609, 0.165810902137309, 0.37615457852371, 0.694700091844425, 0.406369912205264, 0.713788981316611, 0.98604616895318, 0.412632691208273, 0.265657348791137, 0.901944909244776, 0.155186035204679, 0.187537273392081, 0.52104885247536, 0.365703935036436, 0.25453486177139, 0.0841719766613096, 0.628656000364572, 0.0704909595660865, 0.706863556755707, 0.666149195982143, 0.715815547155216, 0.00528890918940306, 0.975355126196519, 0.527055290760472, 0.899172759382054, 0.957532194210216, 0.658752659568563, 0.344424302922562, 0.2046960121952, 0.687126718461514, 0.473313754890114, 0.961187461623922, 0.314407486002892, 0.232060168636963, 0.113535447046161, 0.0349617367610335, 0.221906953491271, 0.358487004879862, 0.499711750773713, 0.0782473210711032, 0.0866039050742984, 0.842941331909969, 0.816047581378371, 0.923963805660605, 0.334405632922426, 0.906712994212285, 0.252333515090868, 0.380540451966226, 0.915626258589327, 0.35415746900253, 0.947123312857002, 0.860210722777992, 0.380404976895079, 0.487789570353925, 0.634832590818405, 0.640068935696036, 0.634012632537633, 0.993126514134929, 0.0175034417770803, 0.402297949651256, 0.368753088638186, 0.515739727998152, 0.520492250332609, 0.543815699405968, 0.350148770958185, 0.746595913078636, 0.599540668539703, 0.547699581133202, 0.3886832697317, 0.641110243741423, 0.622626445256174, 0.919440605677664, 0.734388117212802, 0.234105409588665, 0.163074032636359, 0.982211061520502, 0.286380226258188, 0.295554094249383, 0.399064833531156, 0.367019140627235, 0.814948691288009, 0.450434213271365, 0.221878143027425, 0.537159221945331, 0.478178460150957, 0.944289953447878, 0.315171339549124, 0.653614811599255, 0.0982084718998522, 0.0594024795573205, 0.981553898425773, 0.534016198478639, 0.381773761939257, 0.999840719159693, 0.805214165244251, 0.525673105148599, 0.201928592287004, 0.560543569736183, 0.151676966575906, 0.97655612975359, 0.493179314536974, 0.929666418582201, 0.0966884610243142, 0.370923896320164, 0.857777167344466, 0.969206613721326, 0.912150028161705, 0.204270572634414, 0.977880251826718, 0.997536866692826, 0.987236440647393, 0.984725103247911, 0.436927223112434, 0.4066730318591, 0.273095565615222, 0.117104296805337, 0.735252079088241, 0.297232897020876, 0.266930972691625, 0.036352553172037, 0.218746813479811, 0.701695952331647, 0.768940483452752, 0.773068535141647, 0.229327258886769, 0.343471271684393, 0.34215387608856, 0.116067103808746, 0.940713643794879, 0.149832303868607, 0.512048265896738, 0.0316586466506124, 0.940439412370324, 0.842621399555355, 0.685597195988521, 0.931987346382812, 0.306275623850524, 0.053521960042417, 0.671688183909282, 0.341585077811033, 0.402601063484326, 0.561860419344157, 0.812826893292367, 0.104102940298617, 0.936594962608069, 0.715448102913797, 0.96874461369589, 0.950946625787765, 0.606674495153129, 0.772566882194951, 0.219297355506569, 0.958962354110554, 0.988491268130019, 0.340824448969215, 0.375512527069077, 0.989846135489643, 0.248819167492911, 0.421539797913283, 0.623354649869725, 0.984537645243108, 0.150825430639088, 0.383700815262273, 0.39797759684734, 0.0300255354959518, 0.34387531131506, 0.945791632635519, 0.745817233109847, 0.526468078838661, 0.422032688278705, 0.229149888968095, 0.209856654750183, 0.580809694714844, 0.148491032421589, 0.825200095539913, 0.666955982102081, 0.507394102169201, 0.522209381684661, 0.45091673405841, 0.0740487526636571, 0.732410241151229, 0.979316103970632, 0.338146158726886, 0.125888167647645, 0.34671655157581, 0.546487892279401, 0.0309693478047848, 0.7553245711606, 0.320405951468274, 0.196124750655144, 0.252193527296185, 0.769751683808863, 0.634436414577067, 0.00821652682498097, 0.972301724134013, 0.592002336867154, 0.362276033731177, 0.352050329325721, 0.310756762046367, 0.956710625905544, 0.295033357106149, 0.766379506327212, 0.156855311477557, 0.335554203018546, 0.780070155626163, 0.495119829894975, 0.895198197336867, 0.112168379826471, 0.612126203952357, 0.678170909406617, 0.127194800879806, 0.627528885379434, 0.926240266533569, 0.59984398400411, 0.337448840495199, 0.924397745868191, 0.695778909139335, 0.765509603079408, 0.548942256486043, 0.21962880785577, 0.904767608037218, 0.510478471871465, 0.483342722523957, 0.804720104672015, 0.159144895616919, 0.370970780262724, 0.940040246583521, 0.0298403608612716, 0.610223229741678, 0.267433063825592, 0.972161175683141, 0.221886028302833, 0.253670916194096, 0.142705517821014, 0.984360406640917, 0.739807909820229, 0.440103750443086, 0.04794463631697, 0.486245572566986, 0.635669178795069, 0.440265899989754, 0.104159684618935, 0.00815276824869215, 0.352733612293378, 0.787388806464151, 0.0882745576091111, 0.598628789419308, 0.398151639616117, 0.856624342966825, 0.147823750739917, 0.36234323028475, 0.433976405067369, 0.964560712920502, 0.59880836121738, 0.261151151265949, 0.500286583090201, 0.10624465229921, 0.808460602303967, 0.894579170038924, 0.3564946311526, 0.78519392805174, 0.303632019320503, 0.0742866906803101, 0.387928622541949, 0.26638915319927, 0.342588948784396, 0.392080900026485, 0.204329446190968, 0.329552048584446, 0.160030674887821, 0.381221475312486, 0.50543671939522, 0.490382314426824, 0.937198794446886, 0.696043694857508, 0.75377980642952, 0.978272980079055, 0.430553355487064, 0.864115052856505, 0.140149360988289, 0.889514454407617, 0.853254947345704, 0.0413746607955545, 0.129497469170019, 0.0482965689152479, 0.871361241443083, 0.191046454710886, 0.106878679944202, 0.0128929321654141, 0.279783717589453, 0.808852315647528, 0.652501518372446, 0.807645164430141, 0.0114019939210266, 0.916903602425009, 0.14753962145187, 0.618681674590334, 0.296518967952579, 0.453450528439134, 0.898933904478326, 0.247283065924421, 0.68024740065448, 0.857393784215674, 0.217936357483268, 0.566173854051158, 0.65607752953656, 0.0595100307837129, 0.707232131622732, 0.31261941511184, 0.845636138692498, 0.153854297939688, 0.252888087648898), actual = c(1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"