Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.2406595        0.2411475 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4974633        0.4974633        0.4945138 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.14191436348483, 0.63429051428102, 0.790977486409247, 0.685903182718903, 0.741540393326432, 0.789675819221884, 0.760734924813733, 0.586958679137751, 0.540035970974714, 0.855661581736058, 0.428666038671508, 0.852162116905674, 0.10091713280417, 0.159106745617464, 0.0663658706471324, 0.152580554131418, 0.93240885133855, 0.390443330164999, 0.863048120168969, 0.472564348485321, 0.274058152223006, 0.713736054720357, 0.950043267570436, 0.96466392534785, 0.157091895816848, 0.493337884545326, 0.724811617285013, 0.755860600154847, 0.650538937654346, 0.0357752146665007, 0.288884038105607, 0.707172268303111, 0.663325224770233, 0.396896919934079, 0.422612170455977, 0.78853329597041, 0.965926974546164, 0.587901047198102, 0.937438471475616, 0.880086329765618, 0.872938592685387, 0.759886531857774, 0.984036471927539, 0.235395700670779, 0.624262064462528, 0.986220662714913, 0.573014121735469, 0.715108038857579, 0.554608030710369, 0.0655742581002414, 0.464481884380803, 0.300380717497319, 0.16182561730966, 0.317307770252228, 0.12691887631081, 0.778772418852895, 0.684263854753226, 0.266576830297709, 0.694051613099873, 0.0147743485867977, 0.639977052109316, 0.264476021751761, 0.403717548120767, 0.222952093696222, 0.369697618996724, 0.973358663031831, 0.828645619563758, 0.329091740539297, 0.6844426933676, 0.615169536555186, 0.0355792311020195, 0.30872368812561, 0.438417154829949, 0.348836028715596, 0.410040350165218, 0.900315564358607, 0.327825513435528, 0.492548030801117, 0.210786286974326, 0.42146140197292, 0.345423024846241, 0.326060186140239, 0.0306379033718258, 0.710921881720424, 0.813377707265317, 0.319666549097747, 0.111539044417441, 0.323859995696694, 0.237892367178574, 0.60221258061938, 0.276193760335445, 0.286896728212014, 0.0505237625911832, 0.493846021359786, 0.776519093196839, 0.859980677487329, 0.630967306671664, 0.56852887920104, 0.107141013722867, 0.0190528756938875, 0.887543562334031, 0.825865421444178, 0.94386306242086, 0.873299703933299, 0.760391794377938, 0.562417343026027, 0.411369574023411, 0.958502056309953, 0.520428251707926, 0.965575769310817, 0.643824360100552, 0.728651809506118, 0.261042339494452, 0.0370674931909889, 0.0911629579495639, 0.171918759820983, 0.0453781455289572, 0.877356799552217, 0.329824509564787, 0.499121165368706, 0.987970248097554, 0.234531746245921, 0.815388696268201, 0.781749192625284, 0.404248359845951, 0.579973920248449, 0.709531051339582, 0.785467533860356, 0.489748512627557, 0.387554777320474, 0.673568047117442, 0.808269590837881, 0.85571540845558, 0.172522987704724, 0.0934985200874507, 0.132676843553782, 0.808545229956508, 0.910455463686958, 0.443873355863616, 0.930824321927503, 0.898112385766581, 0.349671817850322, 0.498563710367307, 0.933396828128025, 0.0320596424862742, 0.869289538590237, 0.203360197367147, 0.999608949525282, 0.759439111221582, 0.399281022837386, 0.329179372638464, 0.759877067059278, 0.607054133433849, 0.782528900541365, 0.307595874415711, 0.564164859009907, 0.985263605136424, 0.787281577242538, 0.160353851970285, 0.347086186055094, 0.716147529194131, 0.0898565738461912, 0.675019471673295, 0.421277806395665, 0.300533652305603, 0.436450660694391, 0.492834662320092, 0.0323446139227599, 0.064215944847092, 0.776329126907513, 0.486597180133685, 0.236522131133825, 0.506756417220458, 0.62618598761037, 0.499631664482877, 0.64922202238813, 0.180871106451377, 0.437318401411176, 0.21798683074303, 0.67424087272957, 0.103031588019803, 0.454710765974596, 0.825844132807106, 0.743261610157788, 0.265163619536906, 0.181186801986769, 0.984435575082898, 0.603702322812751, 0.866132882423699, 0.580063133267686, 0.442765210056677, 0.953254983294755, 0.954147563781589, 0.254018806386739, 0.232221879530698, 0.0971849411725998, 0.25480905151926, 0.473572750808671, 0.537502335151657, 0.77337090857327, 0.150149528402835, 0.423674257006496, 0.0787930744700134, 0.525503183016554, 0.22198471124284, 0.193610498216003, 0.750709086190909, 0.0931526215281337, 0.861551888287067, 0.758583191782236, 0.880721391411498, 0.782087008468807, 0.807839270215482, 0.882214418845251, 0.566953647648916, 0.652365655638278, 0.291400991613045, 0.104907061439008, 0.171999318525195, 0.0966587017755955, 0.347573088249192, 0.321839826181531, 0.9525950725656, 0.805439060088247, 0.786307288333774, 0.296371696284041, 0.128610474755988, 0.215127204544842, 0.943978140829131, 0.00581335206516087, 0.0451310437638313, 0.675099122105166, 0.380447756033391, 0.279251410393044, 0.339270147960633, 0.630215655080974, 0.905253741890192, 0.00356314750388265, 0.46503909653984, 0.240248094778508, 0.809573478298262, 0.47263719025068, 0.638793037738651, 0.00770333013497293, 0.282915286719799, 0.525958485901356, 0.652334264479578, 0.117247028276324, 0.164148480631411, 0.35389844677411, 0.956876727519557, 0.0769400547724217, 0.183379801455885, 0.491004107985646, 0.74570837430656, 0.353918864391744, 0.00451352540403605, 0.423443848732859, 0.962527581257746, 0.261395163368434, 0.723176322877407, 0.183452879311517, 0.817621830385178, 0.447323732310906, 0.269215423613787, 0.530985662247986, 0.385416295612231, 0.340732836630195, 0.645948073128238, 0.694945079041645, 0.733680734178051, 0.945426716702059, 0.673545516096056, 0.165580636356026, 0.662034640787169, 0.370233210967854, 0.684992088470608, 0.277744458522648, 0.194812040310353, 0.41842402308248, 0.955181964673102, 0.0758464939426631, 0.167900968808681, 0.264144452288747, 0.686417262069881, 0.956281346734613, 0.302822690457106, 0.0975411648396403, 0.47749491315335, 0.569223360856995, 0.38349073077552, 0.143662026152015, 0.0680115928407758, 0.15652293106541, 0.820624808315188, 0.21192361949943, 0.547678766073659, 0.62769722705707, 0.474094812525436, 0.22194632072933, 0.500197410117835, 0.0675702935550362, 0.466146209975705, 0.48257419327274, 0.678129294188693, 0.467298623872921, 0.50269478186965, 0.776546949753538, 0.0729715856723487, 0.593863838817924, 0.094378849491477, 0.0740661316085607, 0.195100259967148, 0.514767528744414, 0.451712452340871, 0.848880748497322, 0.569850101368502, 0.863738307030872, 0.0632405173964798, 0.272660598158836, 0.02910813735798, 0.956206367816776, 0.0696897855959833, 0.0980333585757762, 0.684780008392408, 0.478093205019832, 0.316730683669448, 0.187117851572111, 0.998502429109067, 0.287219968158752, 0.175767856417224, 0.0526385912671685, 0.801694822730497, 0.784092605113983, 0.0246905016247183, 0.913202526746318, 0.819455747958273, 0.783341530710459, 0.0911866931710392, 0.797086457256228, 0.350821301341057, 0.311916272854432, 0.0287599114235491, 0.0384062603116035, 0.918375109322369, 0.716499706730247, 0.588707371149212, 0.88832320808433, 0.886613108916208, 0.979588659247383, 0.522732285317034, 0.262059952830896, 0.889106964692473, 0.801329009467736, 0.901028379332274, 0.781659075058997, 0.855074935592711, 0.199121966958046, 0.217428604606539, 0.441568672191352, 0.543103039963171, 0.71583079197444, 0.785450605209917, 0.7179298363626, 0.315872736042365, 0.932783390628174, 0.692695069825277, 0.499954864615574, 0.0642976446542889, 0.750287707895041, 0.599082921398804, 0.542626060545444, 0.233402970945463, 0.812668215250596, 0.354595330078155, 0.35756701673381, 0.975320727797225, 0.46066351630725, 0.710144226672128, 0.188049771822989, 0.0437960897106677, 0.614047951996326, 0.589123852550983, 0.383851961931214, 0.0870371065102518, 0.451341135893017, 0.54872981319204, 0.168628480983898, 0.548615786014125, 0.236034645466134, 0.566987765021622, 0.813774429960176, 0.75799166993238, 0.738229739945382, 0.572856480255723, 0.881796087836847, 0.171287353383377, 0.373352361377329, 0.67838527308777, 0.430014922982082, 0.103229262633249, 0.209714665776119, 0.564101732801646, 0.297809056937695, 0.50156359304674, 0.879218545975164, 0.582071995362639, 0.82973123434931, 0.289359825197607, 0.487927591660991, 0.458799398737028, 0.475392546970397, 0.0354454298503697, 0.962048585526645, 0.976298868888989, 0.256482857977971, 0.948916723718867, 0.246075221803039, 0.832223621429875, 0.185239686165005, 0.368940600194037, 0.461285979952663, 0.253588971681893, 0.251304440898821, 0.848929548636079, 0.709831509040669, 0.755733271595091, 0.0129309252370149, 0.410356042440981, 0.999064512550831, 0.852896912023425, 0.884444184834138, 0.873567088274285, 0.154911418445408, 0.780573947355151, 0.934848060831428, 0.274158722022548, 0.336132272379473, 0.71887387521565, 0.892206993885338, 0.0977991151157767, 0.77191948867403, 0.657033230410889, 0.325878756120801, 0.394532937556505, 0.550222428515553, 0.364478996954858, 0.0523148595821112, 0.28103396226652, 0.654461642727256, 0.906059772940353, 0.620544486213475, 0.899854888906702, 0.930095831630751, 0.444958072854206, 0.806008540093899, 0.175950312055647, 0.413000469561666, 0.122229701606557, 0.326127413194627, 0.473269533133134, 0.453777363989502, 0.89616581518203, 0.198288318235427, 0.624185404274613, 0.721806879620999, 0.88380074664019, 0.384298644959927, 0.523842883529142, 0.961072626058012, 0.648445819504559, 0.501959494780749, 0.823942131362855, 0.864090667106211, 0.42547728959471, 0.426554112927988, 0.446538306074217, 0.650845696916804, 0.32420647284016, 0.934730023145676, 0.273406781256199, 0.119741136906669, 0.583885692758486, 0.762941397028044, 0.84590566623956, 0.336873052408919, 0.568959786556661, 0.106150431092829, 0.944459834368899, 0.536363259190693, 0.383744411403313, 0.380103467963636, 0.733411290915683, 0.146806452190503, 0.234261958161369, 0.0653452097903937, 0.400910546304658, 0.518398380838335, 0.905640497338027, 0.359185043023899, 0.689539575017989, 0.346280854195356, 0.298844408011064, 0.818776284111664, 0.980809113942087, 0.546943640103564, 0.853081209817901, 0.450969616649672, 0.949279041495174, 0.659505604300648, 0.0548248437698931, 0.484709083335474, 0.126295921625569, 0.594108098885044, 0.579518739599735, 0.0511747130658478, 0.623966770712286, 0.171773714246228, 0.274435948813334, 0.24362677638419, 0.898684596875682, 0.387957392958924, 0.843321976950392, 0.439571560593322, 0.865994107211009, 0.438021262641996, 0.640571585623547, 0.928833038546145, 0.18706594593823, 0.610011083073914, 0.483749112812802, 0.219571901950985, 0.271525544580072, 0.49980151001364, 0.888883665204048, 0.52327621798031, 0.00369739951565862, 0.748653950169683, 0.243081799009815, 0.558433457277715, 0.534929154440761, 0.630421467823908, 0.0853311568498611, 0.162602644180879, 0.949273357400671, 0.00345315015874803, 0.275949088856578, 0.304058556910604, 0.148029564414173, 0.318730127997696, 0.136581250699237, 0.790765603072941, 0.72914654458873, 0.0410842976998538, 0.63166583981365, 0.894586878828704, 0.0639995406381786, 0.499447822570801, 0.0641648415476084, 0.336927481228486, 0.897462769178674, 0.286458666902035, 0.928620819235221, 0.330429133027792, 0.727469770470634, 0.777127610752359, 0.0380928665399551, 0.908106942893937, 0.415950732305646, 0.330151484115049, 0.392400078708306, 0.688958306796849, 0.818808801006526, 0.177853393368423, 0.841153896646574, 0.619082408724353, 0.918844782514498, 0.45963366352953, 0.0735992286354303, 0.243214936228469, 0.681217951467261, 0.900919805979356, 0.0658191835973412, 0.564427902456373, 0.340263202087954, 0.581709212623537, 0.394044805318117, 0.249252398032695, 0.758723431034014, 0.372027209494263, 0.0665210010483861, 0.242555917706341, 0.183106787502766, 0.349656477337703, 0.325026563368738, 0.98250416177325, 0.646921197883785, 0.279282179195434, 0.276609755354002, 0.508964504580945, 0.971973360981792, 0.198898227885365, 0.101862834766507, 0.566955457208678, 0.783378395950422, 0.852307451656088, 0.715352854225785, 0.972713008988649, 0.426560759544373, 0.305198201676831, 0.162666606949642, 0.696816319599748, 0.564974162727594, 0.143067353405058, 0.024368331534788, 0.470657623605803, 0.217100539477542, 0.595957240555435, 0.249602454714477, 0.482570367632434, 0.327444799244404, 0.00434515019878745, 0.912858271738514, 0.06678012968041, 0.561606575967744, 0.743281151400879, 0.115419883048162, 0.153977661393583, 0.729426285717636, 0.590711073717102, 0.607144724577665, 0.197227441705763, 0.197560318280011, 0.399394837440923, 0.211888242280111, 0.50348916859366, 0.347514338325709, 0.927542031742632, 0.545147011289373, 0.246482798364013, 0.700394410872832, 0.714197921566665, 0.0410145518835634, 0.717011444736272, 0.502510239602998, 0.348449159879237, 0.683365094941109, 0.183831684524193, 0.751020860625431, 0.931332837091759, 0.552546565420926, 0.0412770591210574, 0.923519812291488, 0.81913488660939, 0.545453266473487, 0.189376584021375, 0.46114941034466, 0.922704559518024, 0.916019012918696, 0.54796979483217, 0.502416067756712, 0.255166214425117, 0.176395944319665, 0.773770829662681, 0.994281020248309, 0.608137198723853, 0.231786682037637, 0.0800741300918162, 0.263140843017027, 0.810497071594, 0.917223618831486, 0.917632043594494, 0.893136333907023, 0.785995732061565, 0.540819897549227, 0.526792867109179, 0.628389159683138, 0.613085685530677, 0.792558224173263, 0.663507976802066, 0.42320506228134, 0.520554394926876, 0.350993529427797, 0.613183557055891, 0.208700997987762, 0.160003131022677, 0.442363786278293, 0.746384773869067, 0.655548724345863, 0.963889083825052, 0.710286759305745, 0.183137700427324, 0.0921166124753654, 0.320280893472955, 0.777994281146675, 0.935717461630702, 0.0279062134213746, 0.815507035702467, 0.195090091787279, 0.713597685331479, 0.0333336263429374, 0.185975107830018, 0.672807497438043, 0.0644338284619153, 0.956671892199665, 0.293735599610955, 0.871029207017273, 0.0373141479212791, 0.941645541228354, 0.743746379157528, 0.662501556333154, 0.249874640023336, 0.588287449907511, 0.503728258889169, 0.164070882601663, 0.544539949158207, 0.792811960913241, 0.363089055987075, 0.13817349052988, 0.214740827213973, 0.411682809703052, 0.279101222986355, 0.567330456338823, 0.768817061558366, 0.382721389178187, 0.320779325207695, 0.492612803122029, 0.856028191745281, 0.12999514513649, 0.150464809965342, 0.237226753029972, 0.935119792586192, 0.00990808173082769, 0.966708632186055, 0.20031488337554, 0.0899602507706732, 0.441106392303482, 0.112925824942067, 0.880418105982244, 0.736926899058744, 0.571267110062763, 0.694957745727152, 0.787706949748099, 0.32125606178306, 0.492666076635942, 0.522476280806586, 0.974279659800231, 0.841674439609051, 0.337896054377779, 0.715619423426688, 0.722032757243142, 0.495109386043623, 0.0602609645575285, 0.255133267957717, 0.762105847708881, 0.344080137088895, 0.887455107877031, 0.521735018119216, 0.954580148216337, 0.396941800136119, 0.464697017101571, 0.0300696135964245, 0.717598179588094, 0.872617213986814, 0.1836042474024, 0.295707644894719, 0.933386084390804, 0.759248530957848, 0.684385779080912, 0.197953235125169, 0.291982156457379, 0.391587608261034, 0.323123715119436, 0.0469341818243265, 0.828540127025917, 0.758198725292459, 0.310950503917411, 0.23223585751839, 0.57795200846158, 0.911027511814609, 0.874137195525691, 0.269865210633725, 0.417780480813235, 0.792478307848796, 0.509848292218521, 0.826296312035993, 0.88185895094648, 0.888116434216499, 0.334566231118515, 0.203092814888805, 0.644425772828981, 0.101084018824622, 0.301480239490047, 0.364646882284433, 0.221227618400007, 0.24864625511691, 0.622034185566008, 0.172970755957067, 0.731391999404877, 0.842174148419872, 0.632068070815876, 0.999974957667291, 0.75828599371016, 0.436278617242351, 0.445558398263529, 0.824087380897254, 0.497876421315596, 0.49555932264775, 0.0162498389836401, 0.341432862216607, 0.52837240183726, 0.0398295209743083, 0.319393675075844, 0.291902075521648, 0.190813926747069, 0.529096369398758, 0.112815481843427, 0.798001935705543, 0.510886074742302, 0.765191806480289, 0.104687700048089, 0.108888386050239, 0.740933498833328, 0.258917858591303, 0.951268019154668, 0.0624788501299918, 0.340809799265116, 0.759212122997269, 0.890250422758982, 0.392633920069784, 0.185521996812895, 0.508640311891213, 0.60991541761905, 0.656805528094992, 0.910600013099611, 0.654245120473206, 0.246132087660953, 0.704983409494162, 0.492111931554973, 0.6157206883654, 0.665794565342367, 0.80577141046524, 0.0340750995092094, 0.591440180083737, 0.73249421804212, 0.522275466006249, 0.435103831114247, 0.365914913360029, 0.687209483934566, 0.600923759862781, 0.240723007824272, 0.178568355739117, 0.469304441707209, 0.0996750795748085, 0.60315345460549, 0.167143699713051, 0.499013786204159, 0.361794788390398, 0.857151990057901, 0.291797031881288, 0.020904803648591, 0.942060890840366, 0.815556461922824, 0.0346737178042531, 0.33482582657598, 0.0573447761125863, 0.509106484241784, 0.42774069821462, 0.65489408839494, 0.794278416084126, 0.504817892564461, 0.865682652452961, 0.304451402043924, 0.0539393336512148, 0.416017265990376, 0.280527503462508, 0.356901864288375, 0.277699374593794, 0.0515514924190938, 0.231860266299918, 0.356808270560578, 0.19813472754322, 0.094163915142417, 0.486471248790622, 0.36327159893699, 0.902451787376776, 0.0829099365510046, 0.657011656556278, 0.293499171733856, 0.4538079185877, 0.440298105822876, 0.729641865706071, 0.998373705660924, 0.880755072692409, 0.091047533089295, 0.36921751475893, 0.268265024991706, 0.683221740182489, 0.93819249747321, 0.144619876053184, 0.867712522158399, 0.707930103642866, 0.989549704361707, 0.373595177661628, 0.710085159866139, 0.47128644073382, 0.367185137001798, 0.850643633399159, 0.7427199233789, 0.803005170309916, 0.491286811418831, 0.465266169281676, 0.304990498814732, 0.436053456971422, 0.973862586542964, 0.372860989999026, 0.14508407865651, 0.0782765771728009, 0.830217688111588, 0.90846159751527, 0.390317936195061, 0.840314609231427, 0.323787768837065, 0.822906853165478, 0.24690629937686, 0.974978712853044, 0.347552214050665, 0.669945081928745, 0.0318949031643569, 0.48624043376185, 0.707021388923749, 0.468970824265853, 0.182749321451411, 0.814991599181667, 0.282747467746958, 0.341748869977891, 0.318680608645082, 0.00487640942446887, 0.482200701255351, 0.751280683558434, 0.589888155926019, 0.405677732778713, 0.564264613203704, 0.891458679456264, 0.886948036262766, 0.815601192880422, 0.375870229443535, 0.822687086649239, 0.558057178510353, 0.481149717932567, 0.967179985716939, 0.869193030754104, 0.359427081653848, 0.0447769348975271, 0.812353460351005, 0.0829560314305127, 0.775837229099125, 0.603930220240727, 0.501614257926121, 0.336400544270873, 0.615031336201355, 0.561319078085944, 0.775463086087257, 0.503751078154892, 0.709684908622876, 0.00525895110331476, 0.277185288257897, 0.448280091397464, 0.0160670452751219, 0.0746992845088243, 0.654821706470102, 0.554249068489298, 0.689288593130186, 0.82410971284844, 0.207379321800545, 0.258567706681788, 0.813403096515685, 0.584821035386994, 0.608881420688704, 0.220878164982423, 0.622828574152663, 0.817511204862967, 0.234615955501795, 0.481347494293004, 0.304691239260137, 0.709337768144906, 0.563218829687685, 0.678017762955278, 0.304817923810333, 0.269155739108101, 0.665280841523781, 0.535295795183629, 0.164996091509238, 0.894171349238604, 0.122146327281371, 0.332904785638675, 0.920505729038268, 0.260192533023655, 0.445183167466894, 0.516139344545081), actual = c(1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.978279913775623, 0.731225847732276, 0.484823375474662, 0.15082635730505, 0.278312311507761, 0.259684837888926, 0.483796955551952, 0.447743197903037, 0.252408315660432, 0.900913658551872, 0.228275200352073, 0.114051779964939, 0.007074442692101, 0.882992878323421, 0.109162660082802, 0.863680903567001, 0.730506794760004, 0.374225333333015, 0.925251849694178, 0.716591222444549, 0.470758054871112, 0.242230461211875, 0.541314311092719, 0.558821648592129, 0.3008106444031, 0.366040860768408, 0.879163671983406, 0.694834198337048, 0.593775569228455, 0.242538331076503, 0.0642086672596633, 0.801176152657717, 0.369942343328148, 0.639841412426904, 0.391188728390262, 0.0917611559852958, 0.40470642503351, 0.714069937588647, 0.878199905622751, 0.784705485682935, 0.569470503367484, 0.433728379895911, 0.680301627144217, 0.708718164591119, 0.87793458881788, 0.995475934585556, 0.408319384092465, 0.634338290896267, 0.616856809938326, 0.0252308216877282, 0.434237185865641, 0.695484001655132, 0.0301183338742703, 0.820773524232209, 0.384294805349782, 0.744234575890005, 0.455377840204164, 0.143491398543119, 0.527967830188572, 0.902015203610063, 0.921647223643959, 0.394716060021892, 0.526796035002917, 0.189887854736298, 0.00663788896054029, 0.363777892431244, 0.431552467867732, 0.629292553290725, 0.604446022538468, 0.22395086218603, 0.202920153271407, 0.490003275452182, 0.842955402797088, 0.462976326001808, 0.950029397848994, 0.913262814516202, 0.200763096101582, 0.128568168729544, 0.173720389371738, 0.543658783426508, 0.260073277866468, 0.150603640591726, 0.882821353152394, 0.398532055318356, 0.5067390401382, 0.874056961620227, 0.335938071832061, 0.984567464562133, 0.934714664239436, 0.358380120713264, 0.317207337124273, 0.798062809510157, 0.934238673420623, 0.854439424816519, 0.0956098225433379, 0.362754742382094, 0.373472362523898, 0.00415622675791383, 0.181089607998729, 0.438726400956511, 0.164842898957431, 0.806544579099864, 0.184234702726826, 0.205573904095218, 0.446206217631698, 0.521558111766353, 0.988801478873938, 0.16166485985741, 0.900668416172266, 0.264766793465242, 0.924781840760261, 0.812270782422274, 0.90469683194533, 0.857751538045704, 0.706149381352589, 0.561572153586894, 0.679677218198776, 0.0769508152734488, 0.132024602266029, 0.400848750956357, 0.0177627357188612, 0.315411403542385, 0.114100072300062, 0.589708607876673, 0.313527612015605, 0.714140386553481, 0.0985103400889784, 0.492643559817225, 0.676358404569328, 0.0827708747237921, 0.954103527357802, 0.694812873611227, 0.371558838291094, 0.0904311565682292, 0.375475749373436, 0.162898086942732, 0.576961438870057, 0.865225208457559, 0.497796242358163, 0.88543467503041, 0.135372876888141, 0.744139957940206, 0.743461876176298, 0.122530269436538, 0.892427243757993, 0.376715257065371, 0.318125648424029, 0.426004647975788, 0.687868720386177, 0.661744950804859, 0.183931664098054, 0.86230328050442, 0.328827558783814, 0.89126412733458, 0.930827661417425, 0.709553576540202, 0.270898466696963, 0.919044250855222, 0.0957899228669703, 0.279079271713272, 0.0713674533180892, 0.0559034277684987, 0.569944184273481, 0.952181115513667, 0.576445021433756, 0.57368729962036, 0.821192819392309, 0.376317935995758, 0.930184450466186, 0.528493191115558, 0.161709127016366, 0.17804519366473, 0.772072865394875, 0.628505457891151, 0.770061775576323, 0.376690004486591, 0.623751681763679, 0.879169053398073, 0.943781612906605, 0.359062942210585, 0.1861901441589, 0.412616331595927, 0.0748233492486179, 0.14799079997465, 0.0228035908658057, 0.257963316980749, 0.395605936646461, 0.168677806388587, 0.79246328654699, 0.152914944337681, 0.34998756274581, 0.942036018008366, 0.629433723399416, 0.562898653792217, 0.972265369258821, 0.51657382841222, 0.505787511821836, 0.860471815103665, 0.0256920182146132, 0.220446394989267, 0.343438117997721, 0.535397220402956, 0.71313994913362, 0.305574611760676, 0.334656479768455, 0.65807826584205, 0.760839389171451, 0.606174290645868, 0.4243686767295, 0.65707724285312, 0.115937183145434, 0.84655087813735, 0.516882116207853, 0.161919106263667, 0.717800643993542, 0.36154705635272, 0.766195821110159, 0.35910793626681, 0.890799303073436, 0.331514093559235, 0.0577754855621606, 0.482637241249904, 0.758094242075458, 0.88558460236527, 0.607051596045494, 0.460442518815398, 0.438726771157235, 0.242088221246377, 0.496726691955701, 0.202283472055569, 0.0111690368503332, 0.640998351154849, 0.0100472285412252, 0.500009308801964, 0.84194585448131, 0.0733557967469096, 0.440143422223628, 0.279007816221565, 0.709891578182578, 0.465398401720449, 0.447335261385888, 0.972792494809255, 0.171287032775581, 0.0143339685164392, 0.506313270656392, 0.357011217158288, 0.0144643443636596, 0.132027272833511, 0.887478071963415, 0.70052678999491, 0.821497490862384, 0.163564340444282, 0.686204416910186, 0.0666125633288175, 0.343614559154958, 0.354268572758883, 0.931067997589707, 0.750912746181712, 0.76827907981351, 0.0264802079182118, 0.311891715973616, 0.963644180214033, 0.0163294698577374, 0.19350113067776, 0.482373040169477, 0.786134627182037, 0.313910364871845, 0.590611851774156, 0.795254808617756, 0.588062942959368, 0.6853428252507, 0.640160393435508, 0.0995299257338047, 0.971891578286886, 0.750848952913657, 0.429501647828147, 0.625673313392326, 0.0105318150017411, 0.45392028754577, 0.557717425515875, 0.288663156563416, 0.932737039169297, 0.277755277929828, 0.604582912987098, 0.409105568658561, 0.78526448761113, 0.0727492708247155, 0.0500890973489732, 0.237669733120129, 0.940193761605769, 0.96835032873787, 0.0401712039019912, 0.968552429229021, 0.432260807603598, 0.627716877264902, 0.562768047908321, 0.627647244604304, 0.185090034734458, 0.901148648234084, 0.633888599462807, 0.544159823330119, 0.507155645173043, 0.08614073949866, 0.53409024188295, 0.31406696815975, 0.602131116436794, 0.398244291078299, 0.900929954368621, 0.264818935887888, 0.13984057912603, 0.885545942466706, 0.04956986871548, 0.334550196304917, 0.173417319310829, 0.77143173920922, 0.40783926169388, 0.185648356331512, 0.78619015798904, 0.244868148583919, 0.443501739762723, 0.410831851651892, 0.692039053654298, 0.232875205343589, 0.189009899739176, 0.512887486256659, 0.688830023864284, 0.267566956114024, 0.564494331134483, 0.260726617183536, 0.203999101417139, 0.395277007250115, 0.157650488661602, 0.296839019050822, 0.195966740138829, 0.834419391350821, 0.550886631011963, 0.058195160003379, 0.667563585797325, 0.065464046318084, 0.719209725735709, 0.503381440648809, 0.386947038583457, 0.486539303790778, 0.521561517380178, 0.0747695346362889, 0.322122904006392, 0.951987921493128, 0.196901304880157, 0.72099592583254, 0.930672865128145, 0.241923448396847, 0.726215588394552, 0.490851637208834, 0.362926172325388, 0.892938664415851, 0.230664477450773, 0.310657378984615, 0.0463123207446188, 0.177540231961757, 0.535566935781389, 0.0986468347255141, 0.408542823977768, 0.981085172854364, 0.729272276395932, 0.607106896815822, 0.557505474891514, 0.572600023355335, 0.195681248558685, 0.341814124258235, 0.327183513203636, 0.501049929764122, 0.138567061629146, 0.510751262307167, 0.382393593434244, 0.1665908487048, 0.713006183505058, 0.847799432696775, 0.362637808313593, 0.108143923804164, 0.96552868001163, 0.484319254988804, 0.881713233422488, 0.187587589491159, 0.943043364444748, 0.697786817792803, 0.395704355090857, 0.667574860388413, 0.412737350445241, 0.450708683812991, 0.313284771516919, 0.744231787975878, 0.741618605796248, 0.642017914215103, 0.868777601281181, 0.614305639639497, 0.510522449156269, 0.922755509149283, 0.233127945568413, 0.458485078299418, 0.618096415419132, 0.546434911433607, 0.69284276291728, 0.577123986091465, 0.989141658414155, 0.321070926031098, 0.247891242150217, 0.150159624405205, 0.788092498201877, 0.9702385335695, 0.0253355551976711, 0.840858400566503, 0.462598903337494, 0.20873499289155, 0.0762021441478282, 0.933245379477739, 0.731744120130315, 0.733735267771408, 0.547454618150368, 0.931578254560009, 0.35263242572546, 0.361669882200658, 0.238347117090598, 0.737136822426692, 0.689387316117063, 0.303953324444592, 0.615697389468551, 0.281231423374265, 0.0846654979977757, 0.990509607829154, 0.690914851380512, 0.0701740772929043, 0.831359406933188, 0.0781432990916073, 0.0742333861999214, 0.812342864926904, 0.82685738359578, 0.999677239917219, 0.559557739412412, 0.183425776427612, 0.0597583749331534, 0.693648805841804, 0.433947709854692, 0.197252390906215, 0.94585385825485, 0.987336891237646, 0.338416361715645, 0.730706406291574, 0.902827552519739, 0.861450944794342, 0.860614435048774, 0.101606702664867, 0.225291311275214, 0.888351693283767, 0.904909058939666, 0.871146848192438, 0.149569154251367, 0.668178441701457, 0.67726525128819, 0.964247262571007, 0.860112163238227, 0.0508809667080641, 0.618104072054848, 0.977932179812342, 0.0961288807448, 0.350321091711521, 0.832962453365326, 0.638455285690725, 0.672214555554092, 0.892477534012869, 0.789017496397719, 0.0950806576292962, 0.374473131261766, 0.972278003813699, 0.754085165681317, 0.322165783494711, 0.844206737354398, 0.547259089304134, 0.477710576495156, 0.36480543226935, 0.945021633291617, 0.320604524109513, 0.397136212792248, 0.889918089611456, 0.617329599102959, 0.683945134514943, 0.593469823943451, 0.162457349942997, 0.106656933901832, 0.647504324093461, 0.211717955535278, 0.848747143754736, 0.978435358032584, 0.848142008762807, 0.193016830598935, 0.735643622698262, 0.34273528191261, 0.079328722320497, 0.48525477712974, 0.398544724564999, 0.260204061400145, 0.603159415768459, 0.861609816085547, 0.958833371754736, 0.115495950914919, 0.215598632115871, 0.38582641677931, 0.811044724425301, 0.182194879744202, 0.768056667409837, 0.857078040018678, 0.650781959528103, 0.934668384259567, 0.906458472134545, 0.375099797034636, 0.106890282360837, 0.963945113355294, 0.699867898831144, 0.838042817777023, 0.885356308892369, 0.935102820396423, 0.670215281890705, 0.965716469101608, 0.0183011584449559, 0.0323380227200687, 0.173342424444854, 0.471931188832968, 0.544793980196118, 0.784831897821277, 0.0914272789377719, 0.001396669074893, 0.618112400639802, 0.234440600732341, 0.924601334379986, 0.526722887298092, 0.820566591108218, 0.0460797289852053, 0.920591359259561, 0.822307599475607, 0.291409489465877, 0.398381886770949, 0.715789947891608, 0.156204817118123, 0.626167503418401, 0.797763711074367, 0.121176616055891, 0.31193167436868, 0.831034654518589, 0.823069414356723, 0.496034478768706, 0.244538495317101, 0.298825913574547, 0.879993375157937, 0.37755894777365, 0.450916427886114, 0.964591555995867, 0.969131278106943, 0.706286692293361, 0.474592681974173, 0.107413961086422, 0.494155317544937, 0.91705276700668, 0.89883249392733, 0.178020452847704, 0.0381456036120653, 0.764674772275612, 0.340354731073603, 0.804976559476927, 0.156247416976839, 0.0427084641996771, 0.320421811193228, 0.401085159741342, 0.0632575189229101, 0.0383351212367415, 0.63381821825169, 0.569660246605054, 0.516333179082721, 0.159375631483272, 0.418925566365942, 0.407105612102896, 0.325644686585292, 0.855930767487735, 0.933823234401643, 0.261050939094275, 0.374462327454239, 0.601373873883858, 0.0545826372690499, 0.505038190400228, 0.460073766764253, 0.907849657349288, 0.963061739923432, 0.602030678419396, 0.186116334516555, 0.602196422405541, 0.400201848009601, 0.193890878697857, 0.0543680584523827, 0.87517896364443, 0.244083074619994, 0.00772692123427987, 0.448342766612768, 0.439816642319784, 0.935965940356255, 0.941072158282623, 0.954276728909463, 0.258100201608613, 0.36229544901289, 0.465900417650118, 0.10141624044627, 0.234271346358582, 0.1896850427147, 0.489106098422781, 0.753116608830169, 0.885165333515033, 0.618936671409756, 0.766243886435404, 0.155167006654665, 0.102654692251235, 0.843683703802526, 0.156402110354975, 0.384877285687253, 0.0898889610543847, 0.0449542431160808, 0.54043195862323, 0.199549096869305, 0.782637358177453, 0.727900866186246, 0.44623900577426, 0.96775265596807, 0.704538507619873, 0.89427592838183, 0.0489946941379458, 0.0143534876406193, 0.271602869965136, 0.240142096765339, 0.620458758668974, 0.779162642778829, 0.0135693557094783, 0.64622634393163, 0.433151526609436, 0.419412111863494, 0.210958967916667, 0.28523495234549, 0.553926510037854, 0.0385173263493925, 0.908324937103316, 0.832007595570758, 0.475169505923986, 0.0653574937023222, 0.000483580632135272, 0.7594236463774, 0.10907775326632, 0.902691175695509, 0.66674452344887, 0.841279887594283, 0.375310136703774, 0.440292937215418, 0.364668838912621, 0.40488748671487, 0.333994102664292, 0.675815984373912, 0.483039944898337, 0.142058807890862, 0.904651791090146, 0.75506354845129, 0.76397942728363, 0.847826689481735, 0.814011369133368, 0.199815271655098, 0.635170171968639, 0.918749373871833, 0.345587442629039, 0.33338700979948, 0.92153747449629, 0.14833528897725, 0.49294633930549, 0.289608637336642, 0.297331260051578, 0.272315996466205, 0.079269232461229, 0.745387923670933, 0.757945766905323, 0.512782976031303, 0.597186446422711, 0.0468004974536598, 0.597731550224125, 0.969533658819273, 0.915133626433089, 0.434926687739789, 0.323716102167964, 0.778233604505658, 0.253292831126601, 0.980462468694896, 0.257857520598918, 0.686724140075967, 0.103097860235721, 0.826663327403367, 0.499450551113114, 0.30934832803905, 0.0506423970218748, 0.378192459931597, 0.576704454375431, 0.601327339885756, 0.125420424621552, 0.728617133572698, 0.25245674001053, 0.620457658777013, 0.308018156560138, 0.418641537660733, 0.0642730426043272, 0.446416325867176, 0.185274890856817, 0.623682762961835, 0.131192601518705, 0.0446413205936551, 0.341269789030775, 0.0477794986218214, 0.849097815807909, 0.940313309896737, 0.592355886474252, 0.098729801364243, 0.140976208727807, 0.520175283076242, 0.604431889252737, 0.235392773058265, 0.59257150371559, 0.373761287657544, 0.362130911322311, 0.713543785270303, 0.0484032628592104, 0.836306308861822, 0.446449619717896, 0.622692828765139, 0.0675803213380277, 0.237904642010108, 0.252099054167047, 0.788714000722393, 0.978966075228527, 0.642930748406798, 0.657712774351239, 0.918167645577341, 0.00646632886491716, 0.814356431365013, 0.562432497274131, 0.256736261071637, 0.714830190641806, 0.485938450554386, 0.679807817097753, 0.801278894068673, 0.827760989079252, 0.301149122184142, 0.292197176022455, 0.909573972225189, 0.357042810181156, 0.610180656192824, 0.857264311751351, 0.730045709293336, 0.202827319037169, 0.358679241035134, 0.347295637940988, 0.666278847493231, 0.0112418155185878, 0.347223692340776, 0.931262092897668, 0.00411120592616498, 0.652517616050318, 0.803674133261666, 0.337999545969069, 0.591789295664057, 0.499104450223967, 0.966555088991299, 0.652867645258084, 0.273628119146451, 0.7083518744912, 0.379901964683086, 0.482649652985856, 0.407675491645932, 0.10172199527733, 0.188172857277095, 0.245737358694896, 0.764266754500568, 0.862545993644744, 0.532444823300466, 0.0962378878612071, 0.93394719180651, 0.0721551179885864, 0.75034533534199, 0.802776807686314, 0.257657482288778, 0.136518556857482, 0.228609618265182, 0.0758776848670095, 0.377196847693995, 0.177272146567702, 0.788011454278603, 0.621299755293876, 0.921166287502274, 0.430261108325794, 0.345433823298663, 0.584421686828136, 0.0861354949884117, 0.991289759520441, 0.875276980688795, 0.955042148474604, 0.486197431804612, 0.21353905252181, 0.163991445675492, 0.021316961850971, 0.634049405576661, 0.943665599217638, 0.366995095973834, 0.213936494197696, 0.840446908492595, 0.397909972816706, 0.667136126197875, 0.608374961651862, 0.960510925855488, 0.311675975332037, 0.845480694202706, 0.86726422701031, 0.328217386035249, 0.366643993649632, 0.754184729419649, 0.562996995169669, 0.289998802589253, 0.284088255371898, 0.671814430970699, 0.74506150232628, 0.687905765837058, 0.909625473665074, 0.171172358794138, 0.245754274306819, 0.419827172532678, 0.316686393227428, 0.949221283197403, 0.604329502908513, 0.980265456717461, 0.228305957745761, 0.890889147296548, 0.771724345628172, 0.0640432927757502, 0.543881885707378, 0.533956691389903, 0.306519929086789, 0.935339452931657, 0.672334246803075, 0.92215000744909, 0.477835261495784, 0.318734684726223, 0.52031286759302, 0.220703968312591, 0.760441268794239, 0.940403021406382, 0.546381716616452, 0.230682030785829, 0.324560231762007, 0.413294273428619, 0.675233911722898, 0.0932144687976688, 0.93331835209392, 0.402831113897264, 0.984008844476193, 0.0439979829825461, 0.643388368189335, 0.109127333387733, 0.488068108446896, 0.122344667091966, 0.826933120377362, 0.630722643341869, 0.598713679704815, 0.132165726739913, 0.978728602640331, 0.800088582094759, 0.65736024081707, 0.304327414836735, 0.230945584131405, 0.562931048218161, 0.181673569604754, 0.813946920214221, 0.52795834094286, 0.799811604898423, 0.986088099423796, 0.5584898639936, 0.963579964824021, 0.704716311534867, 0.178533615311608, 0.404334067599848, 0.881225528428331, 0.298826008103788, 0.858395943185315, 0.0235125627368689, 0.977763851638883, 0.567071399418637, 0.873622236773372, 0.269405908184126, 0.279413742013276, 0.268868302926421, 0.293212777934968, 0.112533006584272, 0.439303405582905, 0.298554287990555, 0.58860056870617, 0.188719135709107, 0.422703677089885, 0.344799143495038, 0.178587092319503, 0.405319281853735, 0.15604184102267, 0.0441412872169167, 0.905891981907189, 0.444033106090501, 0.987498230533674, 0.0937763748224825, 0.0376218443270773, 0.809070778544992, 0.507321858080104, 0.627311899326742, 0.424188958248124, 0.66173779242672, 0.145962473237887, 0.414138940162957, 0.994340462610126, 0.83818979235366, 0.0572505476884544, 0.127867789007723, 0.511973434127867, 0.124516997020692, 0.0392074349801987, 0.690051621058956, 0.382618608186021, 0.215704874834046, 0.0650117779150605, 0.428338379599154, 0.306969077326357, 0.893897890578955, 0.762909083859995, 0.0607751472853124, 0.834336726693437, 0.622450248105451, 0.644959775265306, 0.422593743307516, 0.92340735765174, 0.95375657803379, 0.352526172762737, 0.644156777765602, 0.648894852260128, 0.585387632716447, 0.547464868985116, 0.148289057658985, 0.949577321531251, 0.451843846589327, 0.876593015156686, 0.903558095451444, 0.686738251242787, 0.134673926280811, 0.169482985744253, 0.154340643901378, 0.54605859820731, 0.940928579075262, 0.395536869065836, 0.533640750916675, 0.285766946850345, 0.729115078691393, 0.0482963463291526, 0.00628873752430081, 0.407663156744093, 0.50625156564638, 0.557997908676043, 0.271116329822689, 0.0184269975870848, 0.895342753501609, 0.91731621301733, 0.874546836595982, 0.928329393034801, 0.26546406163834, 0.564171185484156, 0.858730383682996, 0.500642935279757, 0.0276154079474509, 0.249787867767736, 0.0506207637954503, 0.593032466014847, 0.530039412900805, 0.360562762944028, 0.433558313874528, 0.621568612987176, 0.924633844289929, 0.999917993322015, 0.336985876318067, 0.675934904254973, 0.327639866620302, 0.166069475701079, 0.323894651373848, 0.698831876041368, 0.277134761679918, 0.198324744822457, 0.0122817237861454, 0.360745730577037, 0.0294869774952531, 0.726408362621441, 0.230482891201973), actual = c(0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"