Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.2662279        0.2692999 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4847505        0.4847505        0.5175815 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.119769197655842, 0.786889323266223, 0.580428196117282, 0.6520256313961, 0.884196049533784, 0.675169856986031, 0.701822288800031, 0.156576246488839, 0.825651101768017, 0.775104302214459, 0.149026017170399, 0.843074785079807, 0.0895986857358366, 0.075163499917835, 0.161459248512983, 0.147193053970113, 0.548052433645353, 0.760969434166327, 0.0922479634173214, 0.620338222943246, 0.527635806240141, 0.141892293235287, 0.981145912548527, 0.780894181225449, 0.716137371491641, 0.864194759866223, 0.0433473943267018, 0.429665485862643, 0.463786243461072, 0.123182726092637, 0.331060833530501, 0.796899899374694, 0.657072151079774, 0.864133286988363, 0.98233911418356, 0.0143304888624698, 0.245212111389264, 0.990812989417464, 0.956130106002092, 0.700101893395185, 0.442274143686518, 0.653838211670518, 0.709718920523301, 0.885009929072112, 0.515296848025173, 0.506547490367666, 0.143742916174233, 0.51141613884829, 0.648786873091012, 0.104660044424236, 0.598236708203331, 0.94061999488622, 0.432902101892978, 0.556725194910541, 0.258457820629701, 0.91884195827879, 0.682375727919862, 0.326584524707869, 0.0693540470674634, 0.150671723531559, 0.415097248507664, 0.97908774157986, 0.730363232782111, 0.35149179212749, 0.127851277124137, 0.336895084008574, 0.453081200830638, 0.710013711592183, 0.0301913833245635, 0.79270269931294, 0.696399055654183, 0.679805669002235, 0.767508279066533, 0.733604791807011, 0.69202805776149, 0.844115476822481, 0.885070406831801, 0.467425781069323, 0.169307353906333, 0.503024380886927, 0.153662010328844, 0.00253058946691453, 0.753163055516779, 0.467267848318443, 0.108233681414276, 0.00922622042708099, 0.775337642757222, 0.27296909224242, 0.310774842975661, 0.513826246839017, 0.700083332369104, 0.00101420190185308, 0.773631231859326, 0.180004818364978, 0.547153555089608, 0.14392208843492, 0.148491983301938, 0.384398865047842, 0.78594058426097, 0.346444201422855, 0.121362535981461, 0.102720947470516, 0.280146370176226, 0.401128395460546, 0.356271995231509, 0.158076540799811, 0.745357065461576, 0.0212995437905192, 0.454746269155294, 0.278470704797655, 0.58887258078903, 0.380397143773735, 0.946388420881703, 0.376074223313481, 0.435289376182482, 0.791132545797154, 0.106655186973512, 0.862481708405539, 0.918938845396042, 0.402660280000418, 0.0171044787857682, 0.943710569990799, 0.775676427176222, 0.978994903853163, 0.52040640451014, 0.993817331735045, 0.562618974363431, 0.300925962626934, 0.131116613512859, 0.845074018463492, 0.142257594969124, 0.617735587758943, 0.25554272858426, 0.162410450633615, 0.238632089924067, 0.609096811618656, 0.655967351049185, 0.208676423411816, 0.148908630479127, 0.359939439920709, 0.949181774165481, 0.855383759830147, 0.893517868826166, 0.854498197091743, 0.820516807492822, 0.894480742281303, 0.85611251113005, 0.757153389742598, 0.095297907711938, 0.0426204460673034, 0.589485361706465, 0.689391587860882, 0.915319855790585, 0.812690439634025, 0.809447862673551, 0.816269022179767, 0.231167029822245, 0.613104535266757, 0.437174641527236, 0.977577005047351, 0.35788690391928, 0.13959341077134, 0.692764528095722, 0.743501091143116, 0.591558794258162, 0.876488895853981, 0.953912617405877, 0.850816470803693, 0.46392568689771, 0.624719612533227, 0.391085413051769, 0.198833627393469, 0.152677722042426, 0.824423551093787, 0.980589663377032, 0.530730342725292, 0.13049233914353, 0.466430799104273, 0.278856951743364, 0.264260369120166, 0.28008596599102, 0.660480839200318, 0.173582643736154, 0.116312081227079, 0.101375626400113, 0.596233732299879, 0.504038717132062, 0.856492380611598, 0.837931385263801, 0.840547501342371, 0.02490274165757, 0.650585281662643, 0.599437270313501, 0.745150598464534, 0.506002386566252, 0.670760983368382, 0.108163833385333, 0.835109908832237, 0.896228608675301, 0.448038849979639, 0.846854147966951, 0.733838487649336, 0.359535333933309, 0.156411087140441, 0.858024250715971, 0.510864690179005, 0.89638325991109, 0.231291087111458, 0.824109628563747, 0.564432283630595, 0.954154780134559, 0.689629980828613, 0.177043113624677, 0.284979083808139, 0.106272294418886, 0.184832126367837, 0.0948995281942189, 0.497498273849487, 0.882022383157164, 0.28594491048716, 0.0876830113120377, 0.224444195628166, 0.682917282916605, 0.537427499890327, 0.253167218295857, 0.206748642493039, 0.0921648771036416, 0.524331696331501, 0.293648789171129, 0.600402004085481, 0.545122695388272, 0.510835019638762, 0.371790644247085, 0.626140426611528, 0.801294909790158, 0.16595747275278, 0.0195264655631036, 0.146261383313686, 0.962413556175306, 0.648106495151296, 0.62820072309114, 0.42538063088432, 0.790409137960523, 0.51920962613076, 0.996722932904959, 0.667803295655176, 0.376602974254638, 0.531402800697833, 0.37459559715353, 0.577869992237538, 0.798655263613909, 0.45517824520357, 0.419353156350553, 0.384006423642859, 0.871497720479965, 0.841426741331816, 0.673813083907589, 0.136135735083371, 0.454144550254568, 0.396904444554821, 0.569246492814273, 0.475141369039193, 0.751727863913402, 0.474650298710912, 0.636548226000741, 0.708924456033856, 0.104130498599261, 0.39347261749208, 0.861534661147743, 0.0404196868184954, 0.789945058291778, 0.884680571034551, 0.297832879936323, 0.427296550245956, 0.219291307032108, 0.857541231904179, 0.414233776042238, 0.818226608447731, 0.455768192419782, 0.365751350764185, 0.727920792065561, 0.0258374358527362, 0.337407199898735, 0.722794860368595, 0.607224733103067, 0.791232448769733, 0.707905953750014, 0.994840543251485, 0.440342254471034, 0.976345396600664, 0.180657079210505, 0.658451511291787, 0.42448357748799, 0.421563511015847, 0.931903789518401, 0.880228996276855, 0.14074090635404, 0.102014736272395, 0.4462655861862, 0.776268946705386, 0.0574439230840653, 0.688640403794125, 0.814139476744458, 0.0888790299650282, 0.567749068606645, 0.790708707179874, 0.694923269562423, 0.259367905091494, 0.457981921499595, 0.372475061332807, 0.816518582170829, 0.481591118732467, 0.0003536781296134, 0.0403896402567625, 0.621971229324117, 0.247674715006724, 0.774849323322996, 0.592195634031668, 0.0612031563650817, 0.914654987864196, 0.24237668979913, 0.851120217004791, 0.0454317494295537, 0.154151144670323, 0.909395907539874, 0.0541739000473171, 0.665552778868005, 0.153014634503052, 0.894578341627494, 0.973907439503819, 0.163349703652784, 0.117805561516434, 0.897234608186409, 0.93281691451557, 0.887831261148676, 0.409965605707839, 0.186032082187012, 0.657924967352301, 0.933720331871882, 0.237803461961448, 0.891390847507864, 0.685362977907062, 0.775029939133674, 0.0330704553052783, 0.0844614522065967, 0.327473348239437, 0.31302326801233, 0.798930196790025, 0.158779170364141, 0.909718480659649, 0.974732968956232, 0.422879645600915, 0.250350379385054, 0.700751801254228, 0.890381907811388, 0.387608575168997, 0.812460809014738, 0.801961714867502, 0.11244197958149, 0.107115377439186, 0.731619832804427, 0.170799591345713, 0.229880865430459, 0.520140158711001, 0.709367321571335, 0.190330300014466, 0.21838254458271, 0.44670963822864, 0.818682939745486, 0.560690420679748, 0.808924229815602, 0.0775783963035792, 0.265173023566604, 0.0911525888368487, 0.110416717128828, 0.286210619611666, 0.0932369478978217, 0.781482052989304, 0.406889542937279, 0.143900903640315, 0.737684357911348, 0.905691504012793, 0.642139567295089, 0.0529119386337698, 0.849295321619138, 0.786520446650684, 0.698827419662848, 0.417354304110631, 0.467936318134889, 0.685625058598816, 0.710631695343181, 0.402017974294722, 0.646536902524531, 0.614616159582511, 0.933302166871727, 0.57413417706266, 0.615606787614524, 0.0379486579913646, 0.913995516719297, 0.780213292688131, 0.466935917036608, 0.4383384892717, 0.443859582766891, 0.351884158328176, 0.605055338470265, 0.910687329247594, 0.251403021160513, 0.455435148673132, 0.291520361090079, 0.390225353883579, 0.522795591270551, 0.961243856465444, 0.698369781021029, 0.537038299487904, 0.936185404192656, 0.977401824202389, 0.0108037407044321, 0.863416778855026, 0.914741596905515, 0.0894568858202547, 0.102846026187763, 0.298721791477874, 0.993449508445337, 0.0355804131831974, 0.612378812627867, 0.09024393488653, 0.562802211847156, 0.147983460919932, 0.103358000284061, 0.863042396958917, 0.710055877221748, 0.221535011427477, 0.658525020815432, 0.225069451378658, 0.00481236423365772, 0.948508303845301, 0.343643189407885, 0.628063892945647, 0.14856969169341, 0.755585591541603, 0.814757905900478, 0.796696950448677, 0.0703467996791005, 0.139519946649671, 0.626926328754053, 0.642102230573073, 0.591462323209271, 0.585388340055943, 0.827666895696893, 0.969170900993049, 0.76546868076548, 0.675228155450895, 0.757783514680341, 0.248338327975944, 0.38023523404263, 0.879789828788489, 0.0961480073165148, 0.398671341128647, 0.546571793965995, 0.779028840363026, 0.569137712940574, 0.0818492702674121, 0.767428864259273, 0.406805922975764, 0.0907606750261039, 0.431892407592386, 0.850421447539702, 0.476980651263148, 0.525275375228375, 0.590204701991752, 0.0588075679261237, 0.000126109924167395, 0.835958984447643, 0.568776146508753, 0.408488472923636, 0.83373145153746, 0.295563611667603, 0.928702779579908, 0.410973091144115, 0.019023938337341, 0.845908406656235, 0.879527397453785, 0.278542848303914, 0.0465059508569539, 0.23632229375653, 0.782806767849252, 0.390257897786796, 0.735293803503737, 0.62491124169901, 0.409593635005876, 0.575293147470802, 0.108184507815167, 0.52341238502413, 0.939866382628679, 0.592261392157525, 0.230013460153714, 0.762159381527454, 0.485041666775942, 0.372819687705487, 0.45845850254409, 0.809148595202714, 0.950530101545155, 0.00510871224105358, 0.484703002497554, 0.793873821618035, 0.574339104117826, 0.924180614994839, 0.0166797044221312, 0.689718239475042, 0.986772555625066, 0.958394453395158, 0.669582878705114, 0.4615283831954, 0.562839908525348, 0.866746530169621, 0.539879345567897, 0.190207439707592, 0.605007652426139, 0.661430403124541, 0.320970240747556, 0.846821749117225, 0.914712043246254, 0.229023903375491, 0.433387188008055, 0.823126405477524, 0.898877854226157, 0.824698674259707, 0.352124843979254, 0.132894050329924, 0.609267526539043, 0.891528618521988, 0.602490063523874, 0.301591834751889, 0.651024467078969, 0.561331629985943, 0.538691645022482, 0.638748099561781, 0.233504034113139, 0.637229434214532, 0.320857466664165, 0.681771375238895, 0.802862304029986, 0.424388605868444, 0.893715228419751, 0.130268544424325, 0.443386715836823, 0.702249923720956, 0.0770273206289858, 0.912385666044429, 0.0198994094971567, 0.956929384963587, 0.970970031106845, 0.342175180558115, 0.162680171895772, 0.542104582535103, 0.771383636631072, 0.399101530201733, 0.800094741396606, 0.695205754134804, 0.882591087371111, 0.691545201698318, 0.814498912310228, 0.428267425158992, 0.972791479900479, 0.146210346370935, 0.6215518664103, 0.146685908781365, 0.379623686894774, 0.790780164534226, 0.702163880225271, 0.711768348934129, 0.857454939046875, 0.60865679429844, 0.933801733655855, 0.695365462219343, 0.00432869023643434, 0.5444532034453, 0.405119422823191, 0.503807740984485, 0.605709951138124, 0.376074142055586, 0.398052816977724, 0.85927824373357, 0.62726676161401, 0.300910783233121, 0.208326662424952, 0.149668717756867, 0.199385511456057, 0.123103479621932, 0.232711605727673, 0.840372251812369, 0.440618416294456, 0.599937159335241, 0.67619692441076, 0.17124127689749, 0.715287309605628, 0.645882260752842, 0.422110785730183, 0.0280618395190686, 0.353327241493389, 0.405959991039708, 0.653093321016058, 0.210915818344802, 0.213513601338491, 0.894109869375825, 0.541812602896243, 0.49663859163411, 0.726234463974833, 0.842473205178976, 0.454048245679587, 0.289572062902153, 0.347349911928177, 0.114762924378738, 0.415975260082632, 0.816424859454855, 0.625859859632328, 0.606393596390262, 0.165210384642705, 0.306226029759273, 0.717216374585405, 0.454296636162326, 0.308353801025078, 0.527322486974299, 0.15342979086563, 0.969540821388364, 0.537263761740178, 0.57721241354011, 0.779119150480255, 0.0920622469857335, 0.698845784412697, 0.0974601197522134, 0.264447442255914, 0.613112769788131, 0.576959841651842, 0.0568954211194068, 0.0926113128662109, 0.859356121625751, 0.512882298557088, 0.391254164511338, 0.257349129300565, 0.583799799671397, 0.122779323952273, 0.0573010793887079, 0.2647334265057, 0.192787231877446, 0.743139684200287, 0.908682374283671, 0.846257732715458, 0.821390360128134, 0.102112071122974, 0.167817018693313, 0.597172585316002, 0.633003479335457, 0.216720091644675, 0.164551502792165, 0.938660443527624, 0.895325887715444, 0.273741841316223, 0.850530479103327, 0.451731509994715, 0.539504564367235, 0.0465838683303446, 0.56111533055082, 0.124584035482258, 0.0793193655554205, 0.618839545408264, 0.527341100620106, 0.851467669941485, 0.856826549628749, 0.344102044124156, 0.731949801091105, 0.889013846870512, 0.675665570423007, 0.429334809537977, 0.0403563808649778, 0.798245427198708, 0.628553287358955, 0.443052862538025, 0.524144859751686, 0.525095668388531, 0.313095352612436, 0.469165291404352, 0.350596626289189, 0.228252578061074, 0.701343226013705, 0.812944381730631, 0.183217311277986, 0.85684864455834, 0.588160095736384, 0.624617285793647, 0.787553443340585, 0.754749367013574, 0.0399778459686786, 0.0652951970696449, 0.707857738714665, 0.772880561184138, 0.465601174626499, 0.982337992172688, 0.95440875296481, 0.192857030080631, 0.788849713047966, 0.512246180558577, 0.573064271360636, 0.99846346443519, 0.761920946650207, 0.839884932385758, 0.397180846193805, 0.752278333995491, 0.140928387641907, 0.91164805018343, 0.156256888294592, 0.410166431218386, 0.35827112197876, 0.797650931868702, 0.652641759021208, 0.260559593327343, 0.568046962842345, 0.224023750517517, 0.0194072821177542, 0.138196693966165, 0.97596750408411, 0.790850915480405, 0.559829230187461, 0.391675744671375, 0.112199514638633, 0.707008177181706, 0.828392328694463, 0.944261036580428, 0.963870681123808, 0.783198407385498, 0.0256242898758501, 0.849855811102316, 0.813278186367825, 0.919121002079919, 0.0306794820353389, 0.84791586920619, 0.218318949453533, 0.933426587842405, 0.325494932942092, 0.497174998279661, 0.996057401876897, 0.0607838772702962, 0.757744165835902, 0.328290768433362, 0.119060383411124, 0.882517097517848, 0.941292558563873, 0.223490918520838, 0.820409360574558, 0.819432580843568, 0.019972822861746, 0.91852347785607, 0.90867332671769, 0.61731836036779, 0.146254463121295, 0.36560728866607, 0.186971054645255, 0.940489064436406, 0.866928355069831, 0.869197097374126, 0.2407731867861, 0.7526939609088, 0.686338564846665, 0.364557737018913, 0.248279249062762, 0.310716125182807, 0.949273034697399, 0.547310506459326, 0.130792673211545, 0.954948117258027, 0.769367600092664, 0.325355634093285, 0.0443975403904915, 0.44793538399972, 0.257206166861579, 0.321976134320721, 0.10286528733559, 0.283350055571645, 0.226863270159811, 0.145634420914575, 0.8488086652942, 0.714586997404695, 0.00817304640077055, 0.692121868021786, 0.504013980040327, 0.226847312180325, 0.15118370577693, 0.699158087139949, 0.308421596186236, 0.253242509439588, 0.802915797103196, 0.529736939352006, 0.0450282245874405, 0.961048511089757, 0.490835502045229, 0.354316025273874, 0.184584075585008, 0.0870461391750723, 0.408357770415023, 0.304156912723556, 0.548111360985786, 0.381212496897206, 0.912344478536397, 0.762111691525206, 0.17820384609513, 0.706262946827337, 0.99781611468643, 0.969643434509635, 0.795126534532756, 0.397477888036519, 0.524164060596377, 0.0720355878584087, 0.0570587392430753, 0.83395209745504, 0.207635981729254, 0.0946489803027362, 0.409093752037734, 0.0798143022693694, 0.577870224835351, 0.192191671347246, 0.0541958056855947, 0.027822645381093, 0.272971634520218, 0.179838228039443, 0.482696749269962, 0.960684150457382, 0.879718431737274, 0.230812917929143, 0.494853235781193, 0.31401152885519, 0.269042953383178, 0.712924983585253, 0.151315122144297, 0.98685103119351, 0.463195468531922, 0.1475128098391, 0.407880464335904, 0.451740299584344, 0.764066412579268, 0.913710744353011, 0.750518396263942, 0.734300551237538, 0.0506006670184433, 0.543993720784783, 0.00896364636719227, 0.871928787324578, 0.294713370734826, 0.367624253500253, 0.313193030888215, 0.803818342741579, 0.603534197434783, 0.377531788544729, 0.859008583473042, 0.91598315956071, 0.648165315855294, 0.192163452506065, 0.076164310099557, 0.162014266708866, 0.822982473997399, 0.931881594005972, 0.531122683547437, 0.918798411032185, 0.793638543924317, 0.86392995622009, 0.651993705425411, 0.482463844586164, 0.876642381539568, 0.678274631034583, 0.048489999724552, 0.230472684139386, 0.272997969994321, 0.704543166561052, 0.413349410519004, 0.51915262802504, 0.300380993867293, 0.680599492974579, 0.797210607910529, 0.625973643502221, 0.612576591782272, 0.769942330662161, 0.846511690411717, 0.177697618957609, 0.893937099725008, 0.560139105422422, 0.19527026778087, 0.480209102155641, 0.637316319160163, 0.613086564932019, 0.216699721291661, 0.779205195140094, 0.91938489372842, 0.71567712049, 0.996602328727022, 0.0521375101525337, 0.867821200517938, 0.384034026879817, 0.177919734967873, 0.188585981260985, 0.251979298423976, 0.0340942228212953, 0.819906591670588, 0.965857884846628, 0.472428880399093, 0.75454274402, 0.17780753527768, 0.0378660617861897, 0.449209602782503, 0.214054346084595, 0.971928646089509, 0.681960188550875, 0.752055704360828, 0.503288495587185, 0.138434596126899, 0.317823360208422, 0.958453536964953, 0.460105115314946, 0.901613884838298, 0.170847010798752, 0.319164345739409, 0.429525855230168, 0.183374527841806, 0.746181606547907, 0.397282508900389, 0.076676475815475, 0.0703917795326561, 0.0948217585682869, 0.364533141488209, 0.802417303202674, 0.840784888947383, 0.263272506883368, 0.160077500389889, 0.783837051130831, 0.762139023980126, 0.0421332677360624, 0.534885881468654, 0.652859514812008, 0.732392391422763, 0.365025607636198, 0.367844994645566, 0.373070418136194, 0.988564225146547, 0.444421515567228, 0.803277506027371, 0.361527540488169, 0.803218603599817, 0.294012908125296, 0.620638402178884, 0.814477812265977, 0.745434247888625, 0.0316199604421854, 0.646471481770277, 0.287825454724953, 0.834840351715684, 0.143500654026866, 0.109932786086574, 0.45113944564946, 0.290319406893104, 0.464996642433107, 0.294670321047306, 0.623576768906787, 0.897656833054498, 0.119343498954549, 0.969941704766825, 0.218776149442419, 0.332024591974914, 0.111294721020386, 0.521068442380056, 0.0533806260209531, 0.382225995650515, 0.0958760471548885, 0.861906656064093, 0.0821293867193162, 0.696739044273272, 0.768909195438027, 0.543464061571285, 0.371992378262803, 0.368614064529538, 0.219257723074406, 0.701991834910586, 0.996371943270788, 0.0140428247395903, 0.839558601146564, 0.265872276155278, 0.82399960514158, 0.293525468790904, 0.513910096371546, 0.403557511046529, 0.598750377539545, 0.15753090637736, 0.638811805984005, 0.890446562087163, 0.298549555009231, 0.42255897494033, 0.998934833100066, 0.968595375772566, 0.203657465055585, 0.0935344565659761, 0.938869360135868, 0.0298823048360646, 0.525758897420019), actual = c(0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.285993443801999, 0.143964788876474, 0.0330228409729898, 0.559745047707111, 0.127198508009315, 0.013786185067147, 0.108822505222633, 0.396747389808297, 0.0804955258499831, 0.517450521932915, 0.824160512536764, 0.321555050089955, 0.483474481618032, 0.589450962375849, 0.741847096942365, 0.865160932997242, 0.486981312511489, 0.386549537302926, 0.0201852361205965, 0.437200689455494, 0.308792619267479, 0.779631460085511, 0.956432387465611, 0.0899103430565447, 0.410467976238579, 0.376906263642013, 0.592882350319996, 0.47390675637871, 0.143497489159927, 0.224001764785498, 0.422587062465027, 0.370869674487039, 0.912521253107116, 0.388544051675126, 0.98214228078723, 0.216012020362541, 0.911466792691499, 0.709757003001869, 0.4946817625314, 0.563667193520814, 0.44690736150369, 0.594715992454439, 0.236572401132435, 0.767145634628832, 0.67895509605296, 0.619058793643489, 0.222082775086164, 0.396528175333515, 0.0696671574842185, 0.383971323492005, 0.138682895805687, 0.69785740529187, 0.823704577051103, 0.848292265553027, 0.490030020009726, 0.645762414671481, 0.301775457570329, 0.421416634228081, 0.272506491281092, 0.9635338941589, 0.300112990895286, 0.613897755043581, 0.328342199092731, 0.2472369375173, 0.30364085896872, 0.133557944791391, 0.148968427442014, 0.120482968399301, 0.210403075907379, 0.787752673961222, 0.678232196485624, 0.993770879693329, 0.754420191748068, 0.771747358841822, 0.852642706828192, 0.459007273195311, 0.967781445011497, 0.395694515667856, 0.890830187825486, 0.628293407848105, 0.0641188719309866, 0.437205084133893, 0.433113045757636, 0.172654672991484, 0.843496019253507, 0.993731464957818, 0.181793013354763, 0.265762820607051, 0.1570237390697, 0.486796941142529, 0.81846302235499, 0.703351773554459, 0.544981357408687, 0.136414238018915, 0.637143238680437, 0.830545797711238, 0.0358152540866286, 0.450704332906753, 0.386177642503753, 0.710206276504323, 0.00138350413180888, 0.160705460002646, 0.156215265858918, 0.0493352108169347, 0.0588646065443754, 0.807814365951344, 0.878224476473406, 0.289022862911224, 0.463885357137769, 0.896938535850495, 0.444745291722938, 0.0433890908025205, 0.452755916630849, 0.244216928258538, 0.0144241077359766, 0.903444847092032, 0.0185531962197274, 0.320773811079562, 0.052097326843068, 0.136973252054304, 0.87867723708041, 0.845549344085157, 0.0461919002700597, 0.763526119524613, 0.0262167111504823, 0.727499593049288, 0.26056715217419, 0.489798522554338, 0.165674832649529, 0.453424291685224, 0.737684190040454, 0.693501424742863, 0.400331567274407, 0.27929448382929, 0.982352401828393, 0.36473153764382, 0.415549388388172, 0.760046240407974, 0.457455849973485, 0.974151174305007, 0.730446465080604, 0.614821527153254, 0.492724833777174, 0.315505660371855, 0.245700585190207, 0.627091642934829, 0.670268844114617, 0.242105490062386, 0.225303569342941, 0.539987675147131, 0.816114551154897, 0.615573308197781, 0.0941245032008737, 0.94100783043541, 0.455965390661731, 0.356021794024855, 0.431686354801059, 0.13999334583059, 0.893990163691342, 0.609519976424053, 0.299314770149067, 0.317700662184507, 0.390539531828836, 0.207486722851172, 0.997855115914717, 0.967350218677893, 0.114052612800151, 0.216869460651651, 0.599035734776407, 0.52452723798342, 0.162941353395581, 0.367638993542641, 0.782676035072654, 0.592565843835473, 0.400992069626227, 0.432459403993562, 0.990160889225081, 0.455668911337852, 0.849531753687188, 0.277818297967315, 0.924093111883849, 0.28591839550063, 0.740081193158403, 0.163436324801296, 0.914842338534072, 0.39693141868338, 0.705070561496541, 0.970207284903154, 0.565347409574315, 0.357170639559627, 0.0447812583297491, 0.528430448845029, 0.92136681615375, 0.335499415872619, 0.225409122183919, 0.81891655176878, 0.377026793314144, 0.564235215540975, 0.382068623555824, 0.870818900177255, 0.119056288851425, 0.932636856101453, 0.605372527381405, 0.62212950293906, 0.0200193505734205, 0.634179642191157, 0.147155231563374, 0.929588777245954, 0.562644395278767, 0.296087317401543, 0.830831470899284, 0.244126333389431, 0.985025199363008, 0.76542635820806, 0.456141558010131, 0.67523833247833, 0.185849234927446, 0.240873353788629, 0.379813133273274, 0.577072325628251, 0.603797474876046, 0.600750217214227, 0.683011150220409, 0.494972145883366, 0.0536465579643846, 0.993848990183324, 0.664639079011977, 0.939216818427667, 0.822737894486636, 0.84555923868902, 0.180975522380322, 0.295394999440759, 0.573309138882905, 0.141614887630567, 0.395424193935469, 0.423232631525025, 0.647100006928667, 0.789714587852359, 0.355831260792911, 0.920944926794618, 0.939882870065048, 0.652572356862947, 0.329630336491391, 0.940272874431685, 0.090636590262875, 0.893623403972015, 0.0937675354070961, 0.392233293969184, 0.199758544797078, 0.325951821403578, 0.356032388284802, 0.734675369225442, 0.224541482515633, 0.300479271449149, 0.575189404888079, 0.64604794816114, 0.929872940061614, 0.738511711126193, 0.289344251388684, 0.222869818098843, 0.411131500499323, 0.688883414957672, 0.3080180992838, 0.245805113809183, 0.476766008650884, 0.920373050263152, 0.952238438185304, 0.749150930671021, 0.324772462015972, 0.790545966010541, 0.837083220249042, 0.367770239477977, 0.675947564188391, 0.391536395996809, 0.399309610947967, 0.247930503915995, 0.673576354980469, 0.916190312709659, 0.852151266066357, 0.796331993769854, 0.260610969970003, 0.873281800188124, 0.749996962491423, 0.857598387170583, 0.154702619416639, 0.187672482337803, 0.519665073836222, 0.822427455568686, 0.309482191456482, 0.0111376002896577, 0.0570777805987746, 0.504939732840285, 0.684973879950121, 0.344467817107216, 0.900805380661041, 0.674084708560258, 0.259015865623951, 0.911810717079788, 0.800266497535631, 0.995625807903707, 0.0962612992152572, 0.136950534768403, 0.922528496477753, 0.038285824470222, 0.582224137848243, 0.117514489917085, 0.326525596436113, 0.237653208198026, 0.270911633269861, 0.0198399408254772, 0.486717357067391, 0.645900620613247, 0.860623686574399, 0.705827753758058, 0.570407698163763, 0.9856759137474, 0.72491015587002, 0.585250020027161, 0.897640237119049, 0.193823393899947, 0.887369740754366, 0.308549675857648, 0.48862152849324, 0.523744067410007, 0.617267339490354, 0.405745939351618, 0.535146901383996, 0.378002075478435, 0.128071954939514, 0.190586322452873, 0.689572518691421, 0.591222339542583, 0.511039800941944, 0.418872239068151, 0.714624688960612, 0.0720462722238153, 0.0702543572988361, 0.924428094644099, 0.214314803248271, 0.738325187470764, 0.628838130272925, 0.203890630509704, 0.485485853627324, 0.94058252312243, 0.508576387306675, 0.41927714808844, 0.600261927582324, 0.582893209997565, 0.994256746722385, 0.172520236112177, 0.716967510059476, 0.976222692290321, 0.652350311400369, 0.295599565841258, 0.306622291216627, 0.943360474426299, 0.0858000868465751, 0.118980769068003, 0.78584112576209, 0.692488320637494, 0.720395139884204, 0.0420757334213704, 0.86421228107065, 0.463443262502551, 0.890414543217048, 0.0421120836399496, 0.00671054795384407, 0.0439960435032845, 0.448201581370085, 0.462805547285825, 0.866812599124387, 0.984161705011502, 0.562607988482341, 0.80703065590933, 0.849449801258743, 0.0221119518391788, 0.744591489201412, 0.261654756497592, 0.150432218331844, 0.440783075988293, 0.602973035769537, 0.76059047668241, 0.709184598410502, 0.762046279851347, 0.746592288604006, 0.350014884956181, 0.493275813292712, 0.74251969344914, 0.684243049938232, 0.384750133845955, 0.626305798068643, 0.524198882980272, 0.891725340858102, 0.711620410904288, 0.329992359969765, 0.116479816380888, 0.262943976791576, 0.880592820700258, 0.0546549847349524, 0.693301214138046, 0.540541596477851, 0.05380759970285, 0.50367550784722, 0.74432666040957, 0.851955546531826, 0.379478020127863, 0.801419326569885, 0.465261443518102, 0.983779586851597, 0.675200740108266, 0.568439349299297, 0.520002309931442, 0.087138299131766, 0.998386348597705, 0.823430834105238, 0.856540485750884, 0.867886180989444, 0.96847219998017, 0.837649120250717, 0.826607564464211, 0.179813620634377, 0.651240619597957, 0.33873966964893, 0.484851435991004, 0.00902433902956545, 0.39860004442744, 0.458082906901836, 0.516947431489825, 0.0354570557828993, 0.380807080306113, 0.273180130403489, 0.128990326775238, 0.566971438238397, 0.716704190010205, 0.906031072372571, 0.935013259062544, 0.791643409058452, 0.213138771476224, 0.403577269520611, 0.0412767126690596, 0.65301339305006, 0.491302244830877, 0.594939963659272, 0.0740763938520104, 0.951836368069053, 0.457753002643585, 0.757549285655841, 0.626208520727232, 0.319278841838241, 0.686094640521333, 0.979812816949561, 0.287901747273281, 0.517242409521714, 0.896362851839513, 0.0353104979731143, 0.460053646471351, 0.435670029604807, 0.556098196655512, 0.178507993696257, 0.984995102277026, 0.850766633637249, 0.3635493805632, 0.0534248678013682, 0.481238719308749, 0.5714868651703, 0.574225426884368, 0.559986033244058, 0.905628881882876, 0.417676175013185, 0.626044108066708, 0.213588058948517, 0.17166662402451, 0.014038490364328, 0.24262618063949, 0.942486173240468, 0.296286721015349, 0.199223984498531, 0.0668923426419497, 0.733256595674902, 0.999369794037193, 0.922201591543853, 0.282589707989246, 0.416456670500338, 0.668545400956646, 0.486431465251371, 0.944877299945801, 0.883392619900405, 0.857102827169001, 0.401398439425975, 0.45465049892664, 0.80159480124712, 0.287578064249828, 0.286527172895148, 0.214054445037618, 0.481365167535841, 0.499629419529811, 0.937225574860349, 0.503759377403185, 0.644686183426529, 0.452693956904113, 0.648196357302368, 0.252383664948866, 0.7938147271052, 0.470494498033077, 0.212685939157382, 0.280612678732723, 0.42618207167834, 0.389256390277296, 0.533341525588185, 0.995603532530367, 0.266919289948419, 0.670381543692201, 0.714738918701187, 0.366881959373131, 0.581341466400772, 0.0118809395935386, 0.907763478113338, 0.999206117587164, 0.458143284777179, 0.707134420285001, 0.138967408100143, 0.171874309889972, 0.358771996339783, 0.38751394324936, 0.973961774259806, 0.228212730493397, 0.221120403846726, 0.0830170740373433, 0.382204678840935, 0.283442238578573, 0.470963058294728, 0.251221544574946, 0.706120803719386, 0.562448932090774, 0.220379879930988, 0.185646459693089, 0.338666639523581, 0.459503412013873, 0.0620296753477305, 0.20740931131877, 0.549840099178255, 0.825049282051623, 0.607409225776792, 0.713597824564204, 0.631786420242861, 0.775899416068569, 0.573432881850749, 0.859179414110258, 0.679161172127351, 0.248948948923498, 0.198977371910587, 0.742019642842934, 0.0919001207221299, 0.246366298524663, 0.982345216907561, 0.448928595287725, 0.659825281007215, 0.318974895169958, 0.650887853233144, 0.148741024779156, 0.287226158194244, 0.719976388849318, 0.280396066140383, 0.1692733715754, 0.798322356306016, 0.749503048602492, 0.278344457270578, 0.25878306850791, 0.470618490828201, 0.873882793588564, 0.160679455613717, 0.186383375432342, 0.266727512469515, 0.580778800649568, 0.593771894229576, 0.526864815503359, 0.922798585845158, 0.773063159547746, 0.212522191461176, 0.255982718896121, 0.492057075956836, 0.770196967059746, 0.754831922473386, 0.992558055790141, 0.544670267263427, 0.736880013952032, 0.573555189883336, 0.0589648671448231, 0.460732688428834, 0.53495492041111, 0.464567597024143, 0.263389362255111, 0.623680637218058, 0.691892410395667, 0.282674783375114, 0.154086652444676, 0.132305024890229, 0.0656681049149483, 0.521262228954583, 0.38644473766908, 0.20731065608561, 0.643783280625939, 0.889618992106989, 0.155199684668332, 0.121227914001793, 0.991351649165154, 0.173148044152185, 0.286958812735975, 0.888629181310534, 0.517009500879794, 0.194575940491632, 0.948886622907594, 0.969192817574367, 0.48309231409803, 0.440670172683895, 0.61999296117574, 0.838965030619875, 0.0484020328149199, 0.19769531628117, 0.549345012521371, 0.194197586504743, 0.619579032296315, 0.45771165471524, 0.225824745604768, 0.327904188074172, 0.00784961483441293, 0.00339419953525066, 0.5003642430529, 0.210060890298337, 0.747633489547297, 0.265292696189135, 0.529250297695398, 0.281943002017215, 0.0297094667330384, 0.336623191600665, 0.815797460731119, 0.401434508152306, 0.0810151088517159, 0.665044383844361, 0.783381290733814, 0.214540356304497, 0.916981251444668, 0.160286704543978, 0.0345018540974706, 0.678404306061566, 0.848679352086037, 0.841699447017163, 0.718659794190899, 0.856480633374304, 0.612918456085026, 0.263822050765157, 0.580457869218662, 0.342267198953778, 0.393477360717952, 0.298391281161457, 0.112796186469495, 0.199679162120447, 0.173927078023553, 0.875706274062395, 0.997914187377319, 0.571088530821726, 0.319056715583429, 0.0778038736898452, 0.129180325428024, 0.799485556781292, 0.159382796147838, 0.343864528462291, 0.465858113020658, 0.807346459943801, 0.354520747205243, 0.00652318494394422, 0.437871305039153, 0.0994571587070823, 0.618407343514264, 0.25797445490025, 0.927839996758848, 0.863906932296231, 0.406550894491374, 0.832149358466268, 0.422866407316178, 0.496087893843651, 0.509216462494805, 0.812341441400349, 0.763499193126336, 0.890909295761958, 0.812662495998666, 0.153958539944142, 0.00628974032588303, 0.77376946923323, 0.174541592365131, 0.343488482758403, 0.406927962554619, 0.645238479832187, 0.448978288564831, 0.811767622828484, 0.16653118073009, 0.92137961066328, 0.110318214166909, 0.747204021317884, 0.433563056634739, 0.448911647545174, 0.276516380952671, 0.333016065414995, 0.166920878924429, 0.74405160266906, 0.906742957653478, 0.631539667025208, 0.434520039474592, 0.469001142308116, 0.232036378467456, 0.161102995509282, 0.237046367954463, 0.466978861251846, 0.248666369123384, 0.563289674464613, 0.419805361889303, 0.520044340752065, 0.923322262009606, 0.771447776118293, 0.244318559067324, 0.57304876530543, 0.768101501278579, 0.436569364275783, 0.389401545515284, 0.972981642233208, 0.00155725097283721, 0.940006850054488, 0.194982468383387, 0.681394480634481, 0.0309557071886957, 0.471703797811642, 0.743430661503226, 0.799151013372466, 0.876530252629891, 0.73329973150976, 0.672125432174653, 0.16456559067592, 0.993858563480899, 0.0987870409153402, 0.670844423351809, 0.868143935920671, 0.266591671621427, 0.237895542057231, 0.853893800638616, 0.421560966642573, 0.117235574638471, 0.651354756206274, 0.546623233472928, 0.188292606035247, 0.726771212881431, 0.2935275004711, 0.661017595790327, 0.303908022120595, 0.661995012778789, 0.777706106659025, 0.125242207432166, 0.162559372140095, 0.838173474417999, 0.67713536741212, 0.618794395821169, 0.0874174509663135, 0.19116149074398, 0.460902850842103, 0.774742436595261, 0.107905752258375, 0.5788556819316, 0.722388189984486, 0.00154637079685926, 0.938668203074485, 0.557552342303097, 0.117072861641645, 0.474050410557538, 0.182914207689464, 0.7621744570788, 0.953085035318509, 0.00743916188366711, 0.048100245418027, 0.655729286372662, 0.0789328371174634, 0.0853924597613513, 0.960992703214288, 0.962597604840994, 0.879008283605799, 0.105257382849231, 0.370254374342039, 0.0222106829751283, 0.786374403163791, 0.834950060583651, 0.164238782366738, 0.942042417824268, 0.864181111799553, 0.700661212671548, 0.321832973277196, 0.582669190363958, 0.22495843260549, 0.24205106520094, 0.899673592532054, 0.378504634369165, 0.729621098609641, 0.63338854466565, 0.780813682125881, 0.0591962430626154, 0.158935944316909, 0.920040389755741, 0.115850222995505, 0.753576203947887, 0.264609778765589, 0.859031666768715, 0.704997248016298, 0.00647432031109929, 0.748440450290218, 0.750646098284051, 0.724458263488486, 0.911545553943142, 0.844357062829658, 0.945680325385183, 0.442794682923704, 0.285255649127066, 0.569203558145091, 0.772179554216564, 0.184771864209324, 0.0820724593941122, 0.884684649761766, 0.462289220653474, 0.8678255781997, 0.99685289664194, 0.547928935615346, 0.143877837574109, 0.397859755204991, 0.643642625538632, 0.832120457664132, 0.572101667057723, 0.679271938977763, 0.638887736247852, 0.45658731716685, 0.38928237836808, 0.674674015957862, 0.395517117809504, 0.820478413719684, 0.877814429812133, 0.844074704684317, 0.196908061858267, 0.864348743343726, 0.451228832593188, 0.499675712315366, 0.582214443944395, 0.055290361167863, 0.55549375503324, 0.44418165483512, 0.0240346919745207, 0.28171234135516, 0.813027868978679, 0.287618143251166, 0.238762303255498, 0.456356729380786, 0.997178330551833, 0.259863627376035, 0.478535439120606, 0.708812444005162, 0.749087505275384, 0.958379584597424, 0.209360215580091, 0.896226588869467, 0.345425244886428, 0.727139965631068, 0.072404776699841, 0.0857708991970867, 0.570345007115975, 0.777735937619582, 0.426940985489637, 0.493469369132072, 0.523918998427689, 0.509340252261609, 0.95294042211026, 0.399901798926294, 0.274042016360909, 0.298413198674098, 0.496625482337549, 0.584278903668746, 0.451705806190148, 0.838624339550734, 0.0228766850195825, 0.037420780165121, 0.315711292438209, 0.606237538857386, 0.790542127098888, 0.0894335547927767, 0.625476974528283, 0.611285318387672, 0.714790652040392, 0.0937138362787664, 0.648340807761997, 0.742630459135398, 0.651292301947251, 0.443666792940348, 0.328347479691729, 0.678200477268547, 0.0133027776610106, 0.348200935870409, 0.297370399581268, 0.531472950940952, 0.285119694424793, 0.82027361006476, 0.898006265284494, 0.52400464611128, 0.890419559087604, 0.255898771574721, 0.750703882193193, 0.0406363122165203, 0.880890606204048, 0.685065967030823, 0.0277209945488721, 0.795287488028407, 0.171526831574738, 0.0037680035457015, 0.0230322044808418, 0.0948408981785178, 0.89272850798443, 0.606191091239452, 0.448719426058233, 0.65341586410068, 0.726838162634522, 0.962490062462166, 0.43958916887641, 0.786701451754197, 0.0344800234306604, 0.529237229842693, 0.829115132102743, 0.632988202152774, 0.621016761288047, 0.696931455284357, 0.123911639675498, 0.0843049152754247, 0.726062341127545, 0.969464701600373, 0.685905271209776, 0.417049128562212, 0.0842189921531826, 0.855340725043789, 0.0800764926243573, 0.380545444786549, 0.505514947231859, 0.726720715407282, 0.84107546065934, 0.97335957060568, 0.64759953529574, 0.0748802092857659, 0.984391164267436, 0.162400013068691, 0.306424118811265, 0.78121350123547, 0.0538967936299741, 0.356378503609449, 0.8252231371589, 0.426129270810634, 0.00615312694571912, 0.381012494675815, 0.623523110989481, 0.697449912549928, 0.267077290685847, 0.661470880964771, 0.397660993970931, 0.577969405101612, 0.488871392561123, 0.170847783330828, 0.603721839142963, 0.301259334664792, 0.49203359009698, 0.37014645175077, 0.131879276363179, 0.998050313675776, 0.585620218189433, 0.0930035326164216, 0.596602075733244, 0.796089637791738, 0.802958994172513, 0.642959500895813, 0.930838595610112, 0.641139476094395, 0.298495940165594, 0.487745341844857, 0.463760033017024, 0.442391647724435, 0.120062548667192, 0.494725848780945, 0.117396035930142, 0.168660351075232, 0.783135697711259, 0.429637359688058, 0.392722587101161, 0.724003676092252, 0.714097738498822, 0.984746461035684, 0.974708754336461, 0.209303797921166, 0.71932591451332, 0.868345164461061, 0.723824476124719, 0.238852485548705, 0.465179553255439), actual = c(1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"