Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.6099089        0.6069439 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5810297        0.5551155        0.5222052 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.345469899475574, 0.0640108925290406, 0.719455058453605, 0.012453495291993, 0.149475969141349, 0.176150866551325, 0.12260086927563, 0.967146608512849, 0.648430579109117, 0.821192617528141, 0.512612773571163, 0.346861664671451, 0.934857069049031, 0.389549385523424, 0.941078842850402, 0.801354881376028, 0.922104802913964, 0.241746945073828, 0.474696248304099, 0.747859293362126, 0.216036741388962, 0.82423946284689, 0.814365823287517, 0.749272323679179, 0.114980980055407, 0.10072892694734, 0.808280924102291, 0.919202052988112, 0.0354022446554154, 0.275320405839011, 0.642338089412078, 0.830109748523682, 0.280295851873234, 0.120754432631657, 0.192065514856949, 0.18101253779605, 0.426363018108532, 0.508760515367612, 0.00636034808121622, 0.102725344942883, 0.462366821244359, 0.783024233300239, 0.419980218634009, 0.511334250448272, 0.456809918861836, 0.277535578701645, 0.0804630303755403, 0.107450793730095, 0.220154776470736, 0.946909908438101, 0.0453417762182653, 0.0376036111265421, 0.741451678564772, 0.605826277984306, 0.529159993864596, 0.355653058271855, 0.35333672654815, 0.698174099205062, 0.564779096981511, 0.710967398947105, 0.422055271221325, 0.403584661427885, 0.154840133618563, 0.937133320141584, 0.36623958661221, 0.188626825809479, 0.129589872667566, 0.615823272615671, 0.839702401775867, 0.914762465981767, 0.935740653192624, 0.174647442763671, 0.448094317223877, 0.541359386174008, 0.256843504030257, 0.711748492671177, 0.334294439060614, 0.196227282984182, 0.457150084432214, 0.453015779610723, 0.503266217187047, 0.0380065152421594, 0.982817930867895, 0.504466374171898, 0.00683910842053592, 0.913148458115757, 0.895436481107026, 0.380315474467352, 0.657708721701056, 0.895581354852766, 0.154861588962376, 0.471056433860213, 0.353057683678344, 0.60659202421084, 0.139838841510937, 0.437847278546542, 0.964357869233936, 0.833180771674961, 0.962474486092106, 0.711859430652112, 0.473782218527049, 0.995789736974984, 0.0973692366387695, 0.353089992422611, 0.274852996692061, 0.544595636660233, 0.043811374111101, 0.548533883877099, 0.208463840652257, 0.745335592189804, 0.146996068768203, 0.49002498970367, 0.53559230780229, 0.868460704106838, 0.271623769542202, 0.444422926288098, 0.335304932203144, 0.761657316703349, 0.571376845473424, 0.505650107981637, 0.852741525741294, 0.981034786673263, 0.810953260399401, 0.218423340236768, 0.248393239220604, 0.228236422641203, 0.0164436465129256, 0.561942785279825, 0.00789260258898139, 0.639758435310796, 0.746093967463821, 0.470329702133313, 0.0783555682282895, 0.513174553168938, 0.55739641492255, 0.926718308823183, 0.780267251655459, 0.40251989220269, 0.762040087254718, 0.590702607529238, 0.616255874279886, 0.0328760757111013, 0.21140885585919, 0.070985309779644, 0.836558246053755, 0.777559325797483, 0.130252768518403, 0.750353719573468, 0.892497151391581, 0.304953052662313, 0.37581932800822, 0.522822684142739, 0.0943378251977265, 0.17451223032549, 0.964779561152682, 0.827224546810612, 0.970210953382775, 0.494230212876573, 0.601789972744882, 0.294797106878832, 0.319402916124091, 0.302776915486902, 0.787595511646941, 0.86505768680945, 0.85663580079563, 0.43263606633991, 0.0430137338116765, 0.98276930488646, 0.121519622625783, 0.103854110697284, 0.835366060491651, 0.955254763830453, 0.555359596852213, 0.720714659430087, 0.737016106955707, 0.625105655519292, 0.890639848308638, 0.62976394360885, 0.1707198319491, 0.574828430078924, 0.27907909755595, 0.926438937429339, 0.200696653220803, 0.974405018147081, 0.131114231888205, 0.467020480660722, 0.641383048612624, 0.851503186160699, 0.826592957833782, 0.993607409531251, 0.271857431391254, 0.392592836404219, 0.651414512190968, 0.842883641133085, 0.908716719131917, 0.641646615229547, 0.211405556416139, 0.287600964307785, 0.446087508462369, 0.382314539281651, 0.964947520056739, 0.707044512964785, 0.207248125690967, 0.00521763623692095, 0.294635090744123, 0.20556124788709, 0.57781550148502, 0.539113794686273, 0.900725186802447, 0.175107330083847, 0.744634717470035, 0.827402229653671, 0.882862918777391, 0.404726288514212, 0.931039741262794, 0.298455088865012, 0.288096578093246, 0.326424442930147, 0.711688022129238, 0.912945508724079, 0.863395972643048, 0.202986490214244, 0.822245123330504, 0.639475275296718, 0.939772286685184, 0.92151592951268, 0.831060724100098, 0.319249216234311, 0.730719147482887, 0.160486589185894, 0.559032729128376, 0.90023123845458, 0.875379095319659, 0.704332425026223, 0.711392658762634, 0.84601865359582, 0.206631451612338, 0.940618343185633, 0.634239243343472, 0.26606206339784, 0.61734294029884, 0.634593438589945, 0.792133097071201, 0.0215364743489772, 0.84109304076992, 0.733565761707723, 0.65577717544511, 0.371843109838665, 0.22983559826389, 0.92138823447749, 0.362294871825725, 0.853254979709163, 0.855222311336547, 0.516075775958598, 0.95369145157747, 0.483137482544407, 0.359114633407444, 0.401957621099427, 0.144079924095422, 0.169345064088702, 0.197909348178655, 0.743120689876378, 0.995491109788418, 0.88903146609664, 0.244006097782403, 0.545347423758358, 0.801080473233014, 0.0992587420623749, 0.531712893396616, 0.216044732602313, 0.359852978261188, 0.776537629542872, 0.468107077758759, 0.228704409440979, 0.527566746110097, 0.331937345676124, 0.554403190733865, 0.208300672238693, 0.831996844150126, 0.44948734738864, 0.378272002795711, 0.176302767591551, 0.352740421891212, 0.835603571729735, 0.830688467249274, 0.550996955251321, 0.832606771262363, 0.208307285327464, 0.834057193947956, 0.594658206449822, 0.991955514065921, 0.513008646899834, 0.833102592267096, 0.995495897484943, 0.856966275488958, 0.431261196034029, 0.786444683559239, 0.691174333682284, 0.314482106128708, 0.765601384686306, 0.155224599177018, 0.584086595568806, 0.576154234353453, 0.810541498241946, 0.215768198948354, 0.068381225457415, 0.707292628940195, 0.259626893559471, 0.895745670888573, 0.0550303258933127, 0.00386889185756445, 0.664397268090397, 0.115765081252903, 0.564163252478465, 0.86388061568141, 0.489605723181739, 0.500797491753474, 0.0821540192700922, 0.929873211076483, 0.865838773082942, 0.852132990024984, 0.498786018695682, 0.58993697585538, 0.115457592066377, 0.600828903028741, 0.123903976753354, 0.605875971028581, 0.928594013443217, 0.37244350113906, 0.469490390270948, 0.416595166316256, 0.480210401583463, 0.745580150745809, 0.883895362727344, 0.240081901662052, 0.645745261805132, 0.826330252224579, 0.892537384061143, 0.556274516740814, 0.42281481390819, 0.591208882629871, 0.852334337774664, 0.570462106261402, 0.326558166649193, 0.0969393465202302, 0.30335635249503, 0.696270777843893, 0.981972889974713, 0.600560822989792, 0.471205321140587, 0.659904373344034, 0.411649383371696, 0.916216056793928, 0.910518324468285, 0.0426605821121484, 0.93525094515644, 0.571930838283151, 0.19614595733583, 0.158292023465037, 0.177924006711692, 0.850721307098866, 0.0570843270979822, 0.150707659078762, 0.841823581373319, 0.930774342268705, 0.519485417054966, 0.10775213316083, 0.475961771095172, 0.355017297202721, 0.822569500422105, 0.192767868051305, 0.25366702163592, 0.976672193035483, 0.907549397554249, 0.0633591446094215, 0.481044966960326, 0.368082660483196, 0.777735631680116, 0.332951163407415, 0.944432938005775, 0.445241775130853, 0.753782430198044, 0.977507397765294, 0.801600662292913, 0.384582874597982, 0.601043472765014, 0.381682469742373, 0.343258841661736, 0.266609760932624, 0.790204102639109, 0.0316113312728703, 0.723332522204146, 0.00431045517325401, 0.0315582465846092, 0.409585151355714, 0.75633823312819, 0.0825213508214802, 0.352920657489449, 0.190427763853222, 0.107198018347844, 0.0767585714347661, 0.293849516427144, 0.0718501128721982, 0.141336527187377, 0.795127712888643, 0.38108859793283, 0.699402977246791, 0.788943688152358, 0.606764843221754, 0.689421850256622, 0.169090443756431, 0.836607686942443, 0.892660710262135, 0.875303087523207, 0.0346281111706048, 0.00846860953606665, 0.336020261282101, 0.432468125829473, 0.0206192238256335, 0.997206310974434, 0.948664001422003, 0.175549332052469, 0.886705411132425, 0.952207385795191, 0.00423292396590114, 0.749648452270776, 0.0301956236362457, 0.8315140879713, 0.764365444425493, 0.211937454063445, 0.28407138842158, 0.996679811505601, 0.092940733069554, 0.0949228445533663, 0.204864983214065, 0.118489133194089, 0.801214785315096, 0.664950810372829, 0.809012923156843, 0.184235519962385, 0.365042492514476, 0.93391639739275, 0.42084239772521, 0.630888941930607, 0.0908926853444427, 0.961305621080101, 0.172401337884367, 0.380180634558201, 0.993340299464762, 0.249869373394176, 0.197071930393577, 0.1160732076969, 0.684548317687586, 0.432719567557797, 0.203254060819745, 0.410987592767924, 0.996809626929462, 0.417195955757052, 0.629708944121376, 0.900414519710466, 0.188267590943724, 0.889340311987326, 0.146965912543237, 0.102163263596594, 0.350538656814024, 0.636657783063129, 0.500544415321201, 0.57913859677501, 0.70448605529964, 0.778284571832046, 0.620730248745531, 0.236721850000322, 0.00866660266183317, 0.0599051162134856, 0.60269575077109, 0.609384681796655, 0.581187420291826, 0.278760643210262, 0.157900589285418, 0.842753975652158, 0.076659640064463, 0.40878855669871, 0.901638913433999, 0.70411951537244, 0.518739299848676, 0.964644454419613, 0.486051419284195, 0.0687903857324272, 0.122586847282946, 0.961875218665227, 0.526397038716823, 0.470516248140484, 0.412639637244865, 0.244274705182761, 0.746414590394124, 0.630842754151672, 0.928159952862188, 0.602885579457507, 0.884563375497237, 0.51854338333942, 0.790289488155395, 0.732807501452044, 0.725078098243102, 0.477201219415292, 0.667976449476555, 0.706900334451348, 0.91454879171215, 0.848573988303542, 0.798415295546874, 0.0265178633853793, 0.74817831418477, 0.59768533334136, 0.717272825771943, 0.606845983071253, 0.666984125040472, 0.15345287322998, 0.566605491563678, 0.181544233579189, 0.668891324428841, 0.273646070156246, 0.90616176975891, 0.298977962229401, 0.566535430261865, 0.180909277172759, 0.362127129919827, 0.8074271583464, 0.0498722845222801, 0.847113194875419, 0.29824487818405, 0.72874970571138, 0.910612896783277, 0.0762692322023213, 0.720024238573387, 0.783637755550444, 0.641761953243986, 0.613314339425415, 0.760881521273404, 0.57793609215878, 0.449741120217368, 0.55183884780854, 0.947570119984448, 0.529152384027839, 0.474229346727952, 0.326280163601041, 0.508741709636524, 0.10699911788106, 0.0619814950041473, 0.036213003564626, 0.20785133400932, 0.533016537781805, 0.536878761835396, 0.441534933401272, 0.0219273876864463, 0.554255678784102, 0.0736085756216198, 0.755987409967929, 0.878331491025165, 0.67525510629639, 0.155178274493665, 0.668133717961609, 0.365177711704746, 0.145254810806364, 0.924319885438308, 0.620664460351691, 0.382236775476485, 0.830971555318683, 0.144447717349976, 0.422071078093722, 0.457225603982806, 0.975363337434828, 0.433074031025171, 0.824458844261244, 0.165400991914794, 0.88533339346759, 0.743792778113857, 0.961649179691449, 0.632191124837846, 0.331200729589909, 0.633392014075071, 0.228674145182595, 0.811274702195078, 0.57665781583637, 0.307162166805938, 0.958259179024026, 0.753712846199051, 0.502836453728378, 0.62822488299571, 0.695476570166647, 0.561025213683024, 0.785704000387341, 0.805050554685295, 0.543868000153452, 0.856272764503956, 0.338484297273681, 0.649110794765875, 0.41827572346665, 0.51993840187788, 0.448145744157955, 0.814858706668019, 0.774055985268205, 0.768713823286816, 0.911447888705879, 0.423092552926391, 0.109370092395693, 0.113234440097585, 0.760949618648738, 0.0466552425641567, 0.714240695349872, 0.581634502857924, 0.92925283848308, 0.48402695171535, 0.0850716726854444, 0.0449223329778761, 0.312618511961773, 0.561815285123885, 0.897101752227172, 0.322266006609425, 0.383372467011213, 0.178414026740938, 0.249066246906295, 0.34024346456863, 0.944230633089319, 0.116904600057751, 0.494793694932014, 0.842048744903877, 0.711224280530587, 0.156794728944078, 0.0404492660891265, 0.129689661320299, 0.79041615780443, 0.490511991316453, 0.570062619866803, 0.34921711566858, 0.486630590166897, 0.289667355827987, 0.732835574774072, 0.633743268670514, 0.917137434938923, 0.159134530927986, 0.678711499553174, 0.101579924114048, 0.107068415498361, 0.570231396937743, 0.0550897060893476, 0.446120972046629, 0.947094741975889, 0.599350800737739, 0.552939622430131, 0.692005598684773, 0.248325432185084, 0.523224644362926, 0.178865013178438, 0.860601996537298, 0.894420173252001, 0.984279610449448, 0.564369760686532, 0.830849141580984, 0.219980883644894, 0.565769475884736, 0.453253411920741, 0.564396535046399, 0.214651297777891, 0.191622402751818, 0.386353149777278, 0.425519945798442, 0.331838794052601, 0.17982151103206, 0.114863035036251, 0.486679333262146, 0.736371562117711, 0.536382354563102, 0.594247489003465, 0.552466888912022, 0.661248457850888, 0.715640347450972, 0.143325504148379, 0.689877652330324, 0.638012736337259, 0.422149484977126, 0.728690556483343, 0.889494922012091, 0.492811866337433, 0.149805160705, 0.463093875907362, 0.725054493173957, 0.744871857808903, 0.0827911149244756, 0.828369878232479, 0.888255473226309, 0.775292643345892, 0.46585699939169, 0.248703440651298, 0.290914577664807, 0.683576287468895, 0.563698630547151, 0.687291940441355, 0.0476742188911885, 0.617155982879922, 0.324039360508323, 0.0767292454838753, 0.167654438409954, 0.434029468568042, 0.0383106861263514, 0.967457125894725, 0.934630271978676, 0.798931366065517, 0.465951877646148, 0.645258939359337, 0.425638063112274, 0.438484796555713, 0.588121668668464, 0.899760067462921, 0.109554252820089, 0.625121558783576, 0.666550738736987, 0.0324468167964369, 0.676147822989151, 0.0549287574831396, 0.350075910799205, 0.990838180528954, 0.283329111523926, 0.911119499476627, 0.716077304678038, 0.54770130245015, 0.0182112001348287, 0.899368614889681, 0.80040110880509, 0.639412716962397, 0.579192757140845, 0.809965830296278, 0.150005283532664, 0.811296618310735, 0.909758959431201, 0.688624654198065, 0.32253741286695, 0.718489867867902, 0.733136646216735, 0.161826145835221, 0.140803319402039, 0.441859606886283, 0.985276416875422, 0.118017613654956, 0.491763533558697, 0.273090504808351, 0.843415413051844, 0.049663238460198, 0.479736402165145, 0.425467702327296, 0.150776025373489, 0.17222306272015, 0.47387190326117, 0.741783332778141, 0.0134965286124498, 0.166393129620701, 0.317409747978672, 0.317562585463747, 0.574646790977567, 0.686900387285277, 0.235425899503753, 0.569776306860149, 0.393928543198854, 0.710729686310515, 0.435423391638324, 0.866888621356338, 0.371039637131616, 0.473294833209366, 0.814895437564701, 0.655591350514442, 0.771268835524097, 0.356397535884753, 0.243632572470233, 0.825749897165224, 0.538561637280509, 0.286011970834807, 0.591302220476791, 0.0205842931754887, 0.277460645651445, 0.840008927276358, 0.369461335008964, 0.223326622974128, 0.555643965490162, 0.674187335185707, 0.316840011160821, 0.731705286540091, 0.780575211625546, 0.676084234379232, 0.858518614433706, 0.576310985488817, 0.335227051284164, 0.24848053837195, 0.470759893069044, 0.834746521199122, 0.636372366221622, 0.355750615941361, 0.311938166618347, 0.945445122430101, 0.54288359824568, 0.815676834434271, 0.481796595966443, 0.460529312491417, 0.304922835202888, 0.94642198481597, 0.619380837772042, 0.161650935886428, 0.761177009670064, 0.507170274620876, 0.762765089049935, 0.137645578244701, 0.952251070644706, 0.747928188415244, 0.915261952672154, 0.671045424416661, 0.573884130222723, 0.427278040908277, 0.608447322389111, 0.443866450805217, 0.333221446024254, 0.837920226855204, 0.488515513017774, 0.908624969189987, 0.483785437652841, 0.668680277187377, 0.653731751255691, 0.516320836031809, 0.440177015261725, 0.50848560128361, 0.586562909418717, 0.542696785414591, 0.69908847543411, 0.00443230499513447, 0.0290708995889872, 0.182342539774254, 0.991035510553047, 0.395651182392612, 0.914249633671716, 0.67071386356838, 0.637998723890632, 0.270323358476162, 0.0686094451230019, 0.335854068864137, 0.703539192676544, 0.338922452181578, 0.77405859506689, 0.792230921098962, 0.669292541686445, 0.273755747359246, 0.935174036538228, 0.0733602882828563, 0.40262365154922, 0.991216199938208, 0.579597098985687, 0.631016208790243, 0.397004245081916, 0.593027769122273, 0.911395648261532, 0.781464826082811, 0.507647179765627, 0.239644586807117, 0.0380030521191657, 0.0663183308206499, 0.910105603514239, 0.941066375235096, 0.618514468893409, 0.64721383084543, 0.642568881390616, 0.962833869969472, 0.603741423925385, 0.294374506687745, 0.253311823355034, 0.698849620763212, 0.854378893505782, 0.0258322292938828, 0.663844087393954, 0.424799669999629, 0.667596915969625, 0.436688868328929, 0.708148223347962, 0.72919550864026, 0.0751036715228111, 0.650752287358046, 0.016247597290203, 0.245287282625213, 0.324657205259427, 0.665647123474628, 0.355993582867086, 0.572067007189617, 0.81049668462947, 0.981400520075113, 0.0450261153746396, 0.0585250235162675, 0.094802220351994, 0.303636721335351, 0.457471842179075, 0.0698154682759196, 0.578186020022258, 0.543363888282329, 0.454993144609034, 0.320511969039217, 0.98263344168663, 0.65935647720471, 0.0768404607661068, 0.695505902869627, 0.543778842780739, 0.839697608491406, 0.388198575470597, 0.979561525629833, 0.818807701347396, 0.984006528975442, 0.697408949490637, 0.157311391551048, 0.819457005010918, 0.423897221218795, 0.628380270907655, 0.702235914301127, 0.803181832889095, 0.704389380058274, 0.799455023137853, 0.392604511929676, 0.354312808252871, 0.186820544069633, 0.851301108952612, 0.358363588806242, 0.341046998277307, 0.646903628250584, 0.656172675779089, 0.288396103773266, 0.598293229937553, 0.73167745093815, 0.0946797204669565, 0.143066289834678, 0.970776966772974, 0.728126472095028, 0.429125476628542, 0.499449633061886, 0.273157916031778, 0.6229437449947, 0.45823956746608, 0.229136195033789, 0.923777542077005, 0.291551791131496, 0.440398706356063, 0.596192844677716, 0.259320598328486, 0.387570500839502, 0.70731039205566, 0.595401655882597, 0.0581927576567978, 0.529565568780527, 0.817259102361277, 0.652657101629302, 0.253605776466429, 0.765193523140624, 0.157136372057721, 0.419669612776488, 0.585185430478305, 0.839246145449579, 0.765347608132288, 0.220515464898199, 0.36633187183179, 0.429087518714368, 0.650091091636568, 0.554531205445528, 0.292209059232846, 0.67345087742433, 0.0554588309023529, 0.147076735738665, 0.0420080958865583, 0.168130784993991, 0.109815764939412, 0.293564241146669, 0.161525655072182, 0.391002953285351, 0.272758846636862, 0.680278734536842, 0.478841006290168, 0.584011434810236, 0.274658245965838, 0.661175755318254, 0.551802543923259, 0.110918787773699, 0.523660351755098, 0.484944362659007, 0.879259528359398, 0.853733009425923, 0.866003940114751, 0.421603061957285, 0.372151180403307, 0.807689136825502, 0.468539382796735, 0.816771432757378, 0.136302794795483, 0.412550411652774, 0.422453293809667, 0.0992037819232792, 0.716403036378324, 0.8643131069839, 0.19644890842028, 0.608511769212782, 0.79676559753716, 0.928099546814337), actual = c(0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.95764093613252, 0.00859338534064591, 0.903377820970491, 0.428328904788941, 0.7774763177149, 0.496008223388344, 0.897327238926664, 0.888097780989483, 0.667882433859631, 0.534623776562512, 0.775796118425205, 0.458631916902959, 0.690602702787146, 0.854828597279266, 0.867315215989947, 0.406794548500329, 0.976699282648042, 0.0288079739548266, 0.856736797839403, 0.351005329051986, 0.112910272553563, 0.592401886358857, 0.771898393984884, 0.40739310788922, 0.62200186168775, 0.421084105269983, 0.193489381344989, 0.459907532902434, 0.264333088882267, 0.328424808802083, 0.195605412824079, 0.656961779342964, 0.00843138573691249, 0.558722713496536, 0.0590833746828139, 0.176406675949693, 0.718727958388627, 0.523382750572637, 0.907233880599961, 0.634565403219312, 0.668970262166113, 0.244344957172871, 0.146807890152559, 0.654812189517543, 0.425429655937478, 0.430904363514856, 0.24446451338008, 0.635526962811127, 0.895338630536571, 0.305643743602559, 0.200825901469216, 0.276637085014954, 0.538698462070897, 0.344941357616335, 0.0551216134335846, 0.00688877096399665, 0.458125193603337, 0.812061477452517, 0.508116460870951, 0.0194607900921255, 0.996365075930953, 0.318300827639177, 0.283132904209197, 0.17125603184104, 0.49818228231743, 0.742538906168193, 0.420855513541028, 0.37891472526826, 0.768244131701067, 0.147700772620738, 0.561019187560305, 0.255107427714393, 0.0479913661256433, 0.405659408075735, 0.0951876055914909, 0.3653667003382, 0.430059564299881, 0.47600497934036, 0.307679106015712, 0.679872132604942, 0.851255339337513, 0.511800528736785, 0.489820218412206, 0.241129224188626, 0.828769713174552, 0.259306966559961, 0.433935468317941, 0.752757265465334, 0.886456390609965, 0.535732547286898, 0.49615277745761, 0.635944396257401, 0.790977687109262, 0.545532545307651, 0.430761939147487, 0.2753387670964, 0.773219812661409, 0.0243446445092559, 0.531544049270451, 0.327541969018057, 0.938182525569573, 0.619864830980077, 0.951653201133013, 0.0900648473761976, 0.430163135984913, 0.811645130859688, 0.665050497511402, 0.423899862449616, 0.0717517230659723, 0.26326344977133, 0.549094604561105, 0.194317944347858, 0.45046885474585, 0.052658102940768, 0.125544261885807, 0.249365585623309, 0.258012709207833, 0.283046580851078, 0.623166308039799, 0.318127777427435, 0.944501272169873, 0.56557802320458, 0.917618144536391, 0.673401032341644, 0.775978017831221, 0.80973615590483, 0.367402796400711, 0.616853224113584, 0.240619231713936, 0.0112398094497621, 0.394048642832786, 0.508149436907843, 0.342718495754525, 0.476280150003731, 0.0758116934448481, 0.641263221390545, 0.0219342424534261, 0.120263324584812, 0.181194311473519, 0.171674131182954, 0.0879264890681952, 0.109812358627096, 0.911511245183647, 0.848314312752336, 0.603586782468483, 0.037958278786391, 0.83521358971484, 0.107464305358008, 0.25369282765314, 0.417172004003078, 0.328407261986285, 0.561177706811577, 0.374845108017325, 0.493636539438739, 0.149654636159539, 0.665041845757514, 0.426434661727399, 0.3570118653588, 0.557023427216336, 0.535824672318995, 0.307395874522626, 0.329575846903026, 0.7908377177082, 0.478965974645689, 0.481399894459173, 0.494176890468225, 0.884622076293454, 0.637992127100006, 0.657941506244242, 0.0490290047600865, 0.778356363298371, 0.448027713224292, 0.628616184694692, 0.703701895661652, 0.299189252778888, 0.955938866361976, 0.210036759963259, 0.387787123676389, 0.407542379805818, 0.90831779781729, 0.895853652385995, 0.248475813539699, 0.914367899531499, 0.852084820857272, 0.614549544639885, 0.835430925013497, 0.164559389464557, 0.840137816267088, 0.341018070932478, 0.178054611664265, 0.886009315028787, 0.207181823905557, 0.928288324736059, 0.678618615493178, 0.386572731891647, 0.962672960013151, 0.509354014415294, 0.774661721428856, 0.18978742742911, 0.301860637031496, 0.122336973901838, 0.586585745215416, 0.226196692092344, 0.483278214465827, 0.617055593291298, 0.851650264812633, 0.674934076843783, 0.777042747242376, 0.381338275503367, 0.165659195976332, 0.862073702504858, 0.651816735742614, 0.531765242572874, 0.369763014838099, 0.663710071472451, 0.574144740821794, 0.995897667948157, 0.621965084923431, 0.481750098289922, 0.992082931334153, 0.196090592071414, 0.468719079857692, 0.381226953817531, 0.611897606635466, 0.383823378710076, 0.677441477775574, 0.777731367386878, 0.801461529452354, 0.511070139938965, 0.970279196742922, 0.876650077989325, 0.178381074452773, 0.338686960749328, 0.635747492779046, 0.193537186831236, 0.719126678537577, 0.875591153046116, 0.83063143119216, 0.512614842271432, 0.331044572405517, 0.446773093892261, 0.509970765560865, 0.623327895300463, 0.771331344731152, 0.418665276141837, 0.219416995532811, 0.021597416838631, 0.172733161132783, 0.840751617914066, 0.925820126896724, 0.068634788505733, 0.590014026267454, 0.0392167554236948, 0.187002389226109, 0.904576855478808, 0.635916977655143, 0.848082900512964, 0.146179819945246, 0.477842525346205, 0.420535478275269, 0.859849873930216, 0.0353990204166621, 0.492500649066642, 0.280767138814554, 0.742373122135177, 0.767618706217036, 0.869208271848038, 0.0950217437930405, 0.69767476292327, 0.533800882752985, 0.875868372153491, 0.642864822642878, 0.466108210152015, 0.60423975950107, 0.217667403863743, 0.390745274489745, 0.511266054585576, 0.488798057893291, 0.597117968834937, 0.147382107330486, 0.758185340324417, 0.202610352775082, 0.66302840039134, 0.438621441600844, 0.520170029019937, 0.650983679108322, 0.115080104442313, 0.525775450514629, 0.105592949548736, 0.516457888064906, 0.0506733539514244, 0.17099553020671, 0.494571561925113, 0.319812116911635, 0.703626003582031, 0.878897990332916, 0.866059041349217, 0.509040673729032, 0.618452433263883, 0.82015042193234, 0.693585601635277, 0.641791817266494, 0.658877622336149, 0.30361669510603, 0.671374103985727, 0.423848187783733, 0.66442783549428, 0.135201016906649, 0.352353097870946, 0.831459606764838, 0.786230680765584, 0.225751163670793, 0.962711461819708, 0.366254736669362, 0.535231362562627, 0.819658474065363, 0.896066423272714, 0.23632210935466, 0.64533582678996, 0.198812121991068, 0.549320094287395, 0.816872156923637, 0.977034110110253, 0.539798103272915, 0.00481405761092901, 0.529980131890625, 0.988794832956046, 0.73610319220461, 0.353545461781323, 0.563102362677455, 0.635768092703074, 0.612883026478812, 0.565055160550401, 0.730101577704772, 0.645946290111169, 0.584668027237058, 0.49169088457711, 0.526536478893831, 0.366388699505478, 0.935059181181714, 0.0475535476580262, 0.197307789465412, 0.678581442218274, 0.449701061472297, 0.903220001608133, 0.964792124927044, 0.120980584993958, 0.392843801528215, 0.476807574508712, 0.92831230070442, 0.955931078176945, 0.84896886209026, 0.269136317074299, 0.863180539570749, 0.430208493955433, 0.918575036106631, 0.920025492087007, 0.470018148655072, 0.740271216724068, 0.0462599468883127, 0.911117931827903, 0.799827647628263, 0.817939167842269, 0.0396858910098672, 0.642812353791669, 0.650805920595303, 0.44605686631985, 0.292776124784723, 0.388900084421039, 0.670947563368827, 0.108979196753353, 0.501613877248019, 0.424644908867776, 0.641082489397377, 0.389385602902621, 0.853811622364447, 0.513000146253034, 0.985258469823748, 0.572129098232836, 0.242552111158147, 0.892171738203615, 0.722636327845976, 0.374567558523268, 0.341564759146422, 0.762455158634111, 0.912971734069288, 0.413153500063345, 0.595246976707131, 0.203393730800599, 0.957172865746543, 0.679404854774475, 0.942611203063279, 0.228811272419989, 0.286628778791055, 0.928784731775522, 0.752832845551893, 0.484374695224687, 0.389400366926566, 0.309831468388438, 0.393619391834363, 0.948024781420827, 0.164625771343708, 0.30030477559194, 0.880873427726328, 0.503723089583218, 0.568857870763168, 0.720673421397805, 0.483926068991423, 0.982010965701193, 0.920093721477315, 0.146753414999694, 0.045580196660012, 0.355559554416686, 0.837357971817255, 0.21660574991256, 0.512412543408573, 0.689993974054232, 0.533333786763251, 0.426220315275714, 0.715011474909261, 0.00697808573022485, 0.748891010414809, 0.73906921595335, 0.354472206206992, 0.961947504431009, 0.0235827891156077, 0.705187690909952, 0.637942634522915, 0.0436063969973475, 0.63853642065078, 0.519709249725565, 0.407153556589037, 0.206421945244074, 0.0253259651362896, 0.711804668419063, 0.446403144625947, 0.679058848647401, 0.383980131940916, 0.517962342360988, 0.48036065325141, 0.368967961985618, 0.853144885739312, 0.246387450024486, 0.248869355767965, 0.169379093684256, 0.220168765168637, 0.742060031276196, 0.430546726565808, 0.0696802153252065, 0.438710303511471, 0.538891690084711, 0.36033794679679, 0.106388629414141, 0.693362492835149, 0.0219774371944368, 0.218742470489815, 0.0995222660712898, 0.763106659986079, 0.280219842446968, 0.533558454131708, 0.690573645755649, 0.385854959487915, 0.149607582716271, 0.707197562558576, 0.950756686506793, 0.789712353143841, 0.166051077423617, 0.00632193894125521, 0.725377751747146, 0.451797107467428, 0.852510330500081, 0.80871706828475, 0.28280016942881, 0.155602412298322, 0.375560553511605, 0.270709378644824, 0.638358848867938, 0.410267883446068, 0.472665888722986, 0.624306749086827, 0.940086713526398, 0.541039486881346, 0.513670698041096, 0.619515184313059, 0.812494662823156, 0.415747930528596, 0.273433410329744, 0.575386021984741, 0.765994533896446, 0.149113978957757, 0.469673730200157, 0.854600019752979, 0.599093149648979, 0.0541954855434597, 0.26863476773724, 0.401092809392139, 0.515548373572528, 0.268312934553251, 0.00706711341626942, 0.244428758276626, 0.80727597582154, 0.297133600804955, 0.801081893499941, 0.4157111546956, 0.994597307872027, 0.268552621128038, 0.954870510380715, 0.278538172598928, 0.0349212277214974, 0.359943042043597, 0.572009686147794, 0.0319468318484724, 0.819559996016324, 0.485513754189014, 0.673205159138888, 0.695398116717115, 0.225435812957585, 0.121347631327808, 0.130824018735439, 0.00375981046818197, 0.876989806769416, 0.135244317119941, 0.227221654495224, 0.0182410893030465, 0.628477333346382, 0.248898845165968, 0.132969471858814, 0.856751645449549, 0.0985656764823943, 0.949835717910901, 0.0041540190577507, 0.171482000732794, 0.432922084350139, 0.541932361898944, 0.155440185684711, 0.282799729378894, 0.247566176578403, 0.762352397898212, 0.654617884662002, 0.0604879474267364, 0.803751443745568, 0.271669244859368, 0.947033511241898, 0.751032131491229, 0.43298260611482, 0.685115625150502, 0.211499062366784, 0.458226707531139, 0.654979729792103, 0.269879430066794, 0.870210673892871, 0.535072803497314, 0.537244880571961, 0.239684268832207, 0.501146006863564, 0.134713592473418, 0.992368452949449, 0.56059684837237, 0.361422538990155, 0.848852513357997, 0.694442068925127, 0.865809537936002, 0.379544846946374, 0.399994765641168, 0.0849049112293869, 0.713505853898823, 0.337076765717939, 0.420946161262691, 0.792384364642203, 0.181770066497847, 0.236713868100196, 0.592692938167602, 0.374916114378721, 0.680201607523486, 0.997892574872822, 0.803048011381179, 0.976320700719953, 0.878276316449046, 0.505940044764429, 0.543861794983968, 0.723162290640175, 0.00501820235513151, 0.733105464372784, 0.644697654759511, 0.728928512427956, 0.133659112034366, 0.833533645840362, 0.316279381280765, 0.104817688232288, 0.616486831568182, 0.676337740151212, 0.0907282254192978, 0.735686011612415, 0.251449485309422, 0.0369650954380631, 0.614584484603256, 0.707182578975335, 0.234914767555892, 0.586458519101143, 0.679705482441932, 0.734651876613498, 0.620355460094288, 0.943035609554499, 0.08930938038975, 0.792211114894599, 0.589532295940444, 0.303186222910881, 0.50275282911025, 0.945934406016022, 0.203546405769885, 0.530645514838398, 0.00190327875316143, 0.0100786546245217, 0.585397797171026, 0.0443264239002019, 0.721250426024199, 0.106598668033257, 0.498251382261515, 0.0785250393673778, 0.195707184961066, 0.950719203799963, 0.738052935572341, 0.659121439559385, 0.844309230102226, 0.134882546495646, 0.928779562469572, 0.290463228244334, 0.33873040555045, 0.204668800113723, 0.191863053012639, 0.511167265241966, 0.0491074749734253, 0.623721855925396, 0.511525786947459, 0.283360112458467, 0.979426287813112, 0.464012166950852, 0.0404667316470295, 0.288195884088054, 0.232158333761618, 0.583431971957907, 0.664035609923303, 0.326600219588727, 0.505566260311753, 0.698656033724546, 0.888018630677834, 0.306695448933169, 0.546641616150737, 0.424067154061049, 0.056712772231549, 0.209433105075732, 0.042680035578087, 0.978706275345758, 0.00369791709817946, 0.0139800910837948, 0.746807702351362, 0.745760508114472, 0.909605682361871, 0.415907957125455, 0.951064352877438, 0.0306382100097835, 0.902254133950919, 0.420105646364391, 0.879227304831147, 0.413690998917446, 0.754418526310474, 0.493196826195344, 0.366167890140787, 0.0322040531318635, 0.373342333827168, 0.439110266044736, 0.180982874007896, 0.75715242465958, 0.690729707479477, 0.850740570807829, 0.0252031777054071, 0.501281161792576, 0.723770348122343, 0.432075320510194, 0.264579604845494, 0.0214574739802629, 0.534302184358239, 0.620618524961174, 0.474243710981682, 0.949360514059663, 0.337612143252045, 0.36856616078876, 0.121905138483271, 0.461224673315883, 0.744619314791635, 0.691913930699229, 0.015445345081389, 0.0626594221685082, 0.227883595740423, 0.74449392198585, 0.548678623978049, 0.282369802473113, 0.59085893118754, 0.0301362667232752, 0.586180702783167, 0.66071921098046, 0.521532466169447, 0.360568044241518, 0.0978502607904375, 0.666731800418347, 0.85539029003121, 0.724457052303478, 0.180203780997545, 0.662358395755291, 0.18876337306574, 0.851296917768195, 0.906279443763196, 0.662669157143682, 0.112824062816799, 0.964770044665784, 0.621393012814224, 0.365540778264403, 0.209835184970871, 0.355920429341495, 0.762237827526405, 0.94814527942799, 0.462292941519991, 0.113662588177249, 0.61382027855143, 0.814820292638615, 0.00692274514585733, 0.254655696451664, 0.482101661153138, 0.891787090105936, 0.902519439347088, 0.491981919622049, 0.366458347765729, 0.868212396046147, 0.871270993258804, 0.509676500223577, 0.217101954622194, 0.338122146436945, 0.582478748401627, 0.934140174183995, 0.228869391838089, 0.458972570253536, 0.227126826066524, 0.402635005069897, 0.991350464290008, 0.567295695422217, 0.596873432863504, 0.50454559084028, 0.75318536395207, 0.501628499478102, 0.146886163158342, 0.510412688134238, 0.523593714693561, 0.658885952318087, 0.186488877050579, 0.120531806023791, 0.919299877481535, 0.668702947208658, 0.995243992423639, 0.717617111746222, 0.815317433560267, 0.888214844977483, 0.998749465448782, 0.503547254018486, 0.632558007258922, 0.579909196356311, 0.792402778053656, 0.634346843697131, 0.0720423480961472, 0.469545562751591, 0.713238321710378, 0.776130519807339, 0.284934953786433, 0.180324136978015, 0.342815743293613, 0.400497121969238, 0.102737489389256, 0.850629056803882, 0.227412421256304, 0.441488841315731, 0.203998021548614, 0.30120835499838, 0.461469704750925, 0.25985228875652, 0.0814642137847841, 0.216564449248835, 0.739679763559252, 0.249771774513647, 0.583667805884033, 0.494264723500237, 0.6796408763621, 0.132127688964829, 0.913235899526626, 0.112206107703969, 0.636582736158744, 0.3406263187062, 0.416559064295143, 0.611571055836976, 0.60061848978512, 0.230896974215284, 0.620945370523259, 0.599269098136574, 0.995472381124273, 0.194395664148033, 0.0513591177295893, 0.956894783303142, 0.406553977169096, 0.430956472409889, 0.0548525261692703, 0.661461934680119, 0.967402207199484, 0.131235111039132, 0.60187283041887, 0.588462697807699, 0.594451716169715, 0.946505750296637, 0.176673363428563, 0.659885227913037, 0.515600781654939, 0.0432182450313121, 0.264045849209651, 0.641003060154617, 0.349203878780827, 0.831463451031595, 0.960743270348758, 0.0681723626330495, 0.800540680065751, 0.327531266026199, 0.964191258419305, 0.28940928238444, 0.393269336316735, 0.302959191612899, 0.754519369918853, 0.0335221120622009, 0.715031035244465, 0.618811666732654, 0.645587127190083, 0.979095655260608, 0.510034596314654, 0.728376017417759, 0.174036767333746, 0.470206365920603, 0.21482231025584, 0.281989737180993, 0.196434479672462, 0.325983900343999, 0.810284976381809, 0.55613828310743, 0.490647173253819, 0.0397206763736904, 0.233509817393497, 0.508072550175712, 0.286467338213697, 0.943728467449546, 0.38979289890267, 0.437681933166459, 0.422354699345306, 0.601952640805393, 0.755609456915408, 0.855797459138557, 0.687406444456428, 0.299148957012221, 0.416708322474733, 0.576961046317592, 0.675181896658614, 0.2607581759803, 0.970707544125617, 0.128021764801815, 0.950097970897332, 0.053146906895563, 0.45210022223182, 0.96031531621702, 0.376576888840646, 0.799786131130531, 0.605233591748402, 0.390249773161486, 0.748338404111564, 0.526736110216007, 0.537803585175425, 0.741156806703657, 0.534066392574459, 0.380251732422039, 0.294594412203878, 0.70871083647944, 0.844642720650882, 0.876733407611027, 0.147696146275848, 0.302128709387034, 0.563270163489506, 0.27312927134335, 0.770546462386847, 0.152092290343717, 0.49679528339766, 0.656293815933168, 0.979698996292427, 0.532722268719226, 0.940801385324448, 0.214203807525337, 0.952594731934369, 0.34985642763786, 0.226590509293601, 0.891952993581071, 0.59872129233554, 0.507840959820896, 0.733387945452705, 0.79773607919924, 0.891346214339137, 0.618105059955269, 0.587501209229231, 0.45577514404431, 0.0561242958065122, 0.479752108221874, 0.137920695589855, 0.44226220366545, 0.44433306879364, 0.785493053030223, 0.45927653927356, 0.116344898939133, 0.0150277731008828, 0.10690009011887, 0.922461411217228, 0.978564221877605, 0.222503158031031, 0.0514889841433614, 0.931294495239854, 0.372535876231268, 0.565293112071231, 0.660364838782698, 0.776315797353163, 0.427400602260605, 0.35664771986194, 0.7066064781975, 0.490872242487967, 0.596549338893965, 0.0932883203495294, 0.0027672543656081, 0.912539630895481, 0.826891632284969, 0.523556785658002, 0.21873842808418, 0.087096971925348, 0.159715347690508, 0.381452713394538, 0.684163013473153, 0.523408764507622, 0.0417275174986571, 0.681684475159273, 0.771921321051195, 0.0835872292518616, 0.967156819067895, 0.257407005643472, 0.373099649557844, 0.882413880899549, 0.851434802869335, 0.250897971214727, 0.0561148263514042, 0.194391334662214, 0.683944577584043, 0.148182108299807, 0.64763838169165, 0.00566077372059226, 0.205161706078798, 0.662725900532678, 0.85345440171659, 0.671299415407702, 0.37134564993903, 0.457185577834025, 0.599402712192386, 0.215231168083847, 0.0670102231670171, 0.383703936357051, 0.100047644227743, 0.453075812198222, 0.629078625002876, 0.287358832312748, 0.226768159074709, 0.168215150712058, 0.605977905215696, 0.436214643530548, 0.587672756519169, 0.578121729427949, 0.0196696992497891, 0.225682966178283, 0.806344309356064, 0.318246398121119, 0.27578122750856, 0.570554331876338, 0.450319566531107, 0.00133490562438965, 0.684459296287969, 0.941041423007846, 0.674745586933568, 0.204552128445357, 0.569385988172144, 0.675157643388957, 0.933278516866267, 0.751357985427603, 0.270262819947675, 0.286863909102976, 0.63003377453424, 0.993521000025794), actual = c(1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"