Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.9067439        0.3934863 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4882649        0.4882649        0.5351228 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.692690561525524, 0.671628434909508, 0.829300832701847, 0.971761912340298, 0.0277879580389708, 0.773112192982808, 0.539035455090925, 0.708761678310111, 0.520373887382448, 0.697838762542233, 0.135997146833688, 0.728843834949657, 0.935592462774366, 0.0821632952429354, 0.213564072735608, 0.194616485387087, 0.497843566117808, 0.524718429194763, 0.604667548555881, 0.139822679106146, 0.179595499299467, 0.280714612221345, 0.0995458173565567, 0.964734756620601, 0.605575168970972, 0.227165274089202, 0.992637225659564, 0.642180402297527, 0.431383256567642, 0.526256381301209, 0.766265562502667, 0.133955425582826, 0.833080881508067, 0.20266821468249, 0.808229353744537, 0.527435581199825, 0.360354110598564, 0.0165476803667843, 0.108900681370869, 0.748164617689326, 0.536420041928068, 0.352460198337212, 0.696343602612615, 0.0706087385769933, 0.0832163009326905, 0.973246005596593, 0.94020680594258, 0.0734433550387621, 0.871845797402784, 0.477365614846349, 0.355553389526904, 0.883185749175027, 0.175465288339183, 0.938142866361886, 0.72054009558633, 0.397615450667217, 0.184794451808557, 0.599745630053803, 0.50249063083902, 0.240426868665963, 0.384666368365288, 0.307847527088597, 0.828604572685435, 0.676870553055778, 0.24278778466396, 0.929214228643104, 0.149611260741949, 0.113837101496756, 0.229657733114436, 0.238653904292732, 0.0447845228482038, 0.951574116013944, 0.374900046736002, 0.823819560464472, 0.492349473293871, 0.632089756429195, 0.502229526173323, 0.435479604173452, 0.306015972513705, 0.684566118055955, 0.54333977936767, 0.606610363349319, 0.00492597348056734, 0.0725286481902003, 0.044689821312204, 0.878070152364671, 0.280800637323409, 0.539886340498924, 0.780387887265533, 0.743816636502743, 0.422186196316034, 0.873722398187965, 0.932964812964201, 0.069126098882407, 0.585156999295577, 0.226631585508585, 0.819190093316138, 0.448324915021658, 0.441981942392886, 0.867495096055791, 0.689750680001453, 0.687059343559667, 0.98873857781291, 0.00894109206274152, 0.0771569188218564, 0.782615134026855, 0.869311605580151, 0.858369988156483, 0.458768931683153, 0.229611292714253, 0.245470468420535, 0.867028935579583, 0.438212125329301, 0.099542502546683, 0.581224259454757, 0.496696370653808, 0.695096766809002, 0.0667531306389719, 0.168411381775513, 0.755939408671111, 0.898733203299344, 0.983027358073741, 0.293383745942265, 0.211372954072431, 0.224021166330203, 0.0100453151389956, 0.345254528103396, 0.922187104588374, 0.795572317205369, 0.225067125633359, 0.540308422176167, 0.924360882956535, 0.266113087767735, 0.403547181049362, 0.104192141909152, 0.843673299066722, 0.471408979501575, 0.454317266819999, 0.181249205954373, 0.655353819951415, 0.416897171642631, 0.212484248913825, 0.00152842840179801, 0.11332893371582, 0.55849249730818, 0.943340444238856, 0.49230728065595, 0.692992246244103, 0.836955212289467, 0.586459068115801, 0.00176827702671289, 0.353132012300193, 0.265720310620964, 0.926932259462774, 0.0363753705751151, 0.350996406748891, 0.161619945894927, 0.394777206005529, 0.732947145588696, 0.285850942367688, 0.0338006266392767, 0.44360969029367, 0.417488541686907, 0.337320792023093, 0.91483666235581, 0.316317684249952, 0.761850020848215, 0.399823192739859, 0.180012021446601, 0.260690274881199, 0.173630079953, 0.249194634146988, 0.250256556086242, 0.780130117200315, 0.331555370241404, 0.418553127907217, 0.558824153617024, 0.857060274807736, 0.485200110124424, 0.404313961043954, 0.565537574701011, 0.317047586198896, 0.936195062240586, 0.34920061705634, 0.540090838680044, 0.84549906780012, 0.295887840678915, 0.0199051951058209, 0.570979595882818, 0.167944888351485, 0.753772990312427, 0.615229266695678, 0.653558830032125, 0.956405470846221, 0.802662016125396, 0.382054124260321, 0.838836463401094, 0.926364234648645, 0.10538218007423, 0.221203470369801, 0.692788525717333, 0.459107203874737, 0.316940479911864, 0.474613548256457, 0.916587307816371, 0.281418963102624, 0.0390755017288029, 0.173174881376326, 0.879195536486804, 0.311905275797471, 0.0682158921845257, 0.188485855003819, 0.370721474289894, 0.0669975823257118, 0.835066082654521, 0.27856356697157, 0.00131114851683378, 0.758220902178437, 0.277963009197265, 0.168595694703981, 0.914978966116905, 0.78712868783623, 0.282572431024164, 0.831524127628654, 0.148094248026609, 0.224733705399558, 0.882800900610164, 0.120376193430275, 0.177820983575657, 0.573863122845069, 0.340971607482061, 0.740618097130209, 0.981418440816924, 0.734067719895393, 0.342787770088762, 0.850563937332481, 0.00302004581317306, 0.31735230772756, 0.674623558996245, 0.186278748326004, 0.268991945777088, 0.0326397812459618, 0.597802036907524, 0.894974216353148, 0.757961966330186, 0.132830949965864, 0.652312171878293, 0.224472827743739, 0.613073931774125, 0.373354680836201, 0.6677114746999, 0.769044200191274, 0.532795847626403, 0.394779198337346, 0.783793397946283, 0.924174693180248, 0.61484452476725, 0.752606137888506, 0.54497856204398, 0.510326465824619, 0.731408658437431, 0.825852268142626, 0.551481142174453, 0.0332642272114754, 0.152494759764522, 0.717049575410783, 0.936341146938503, 0.229813841171563, 0.676863700849935, 0.798575313994661, 0.317248010076582, 0.536440843716264, 0.431367475073785, 0.775837881956249, 0.121694604866207, 0.293534624855965, 0.708506355062127, 0.878844649763778, 0.513946482911706, 0.887447231216356, 0.304534343769774, 0.576049239141867, 0.943422847893089, 0.311629044357687, 0.448209371417761, 0.457436098949984, 0.534881311701611, 0.705771701876074, 0.396478082751855, 0.798306816956028, 0.304392605554312, 0.367571371374652, 0.77217966131866, 0.0649613626301289, 0.522255082847551, 0.847068915143609, 0.591005897382274, 0.625659190816805, 0.126663094153628, 0.855351489735767, 0.476749806432053, 0.0738673771265894, 0.513772871810943, 0.0345986974425614, 0.717845350736752, 0.476959283696488, 0.803054822375998, 0.868569550802931, 0.602771695237607, 0.128452498232946, 0.25938228610903, 0.394030438968912, 0.411608170717955, 0.309453904163092, 0.873415064066648, 0.286262245615944, 0.560912935994565, 0.313831062987447, 0.523158641066402, 0.00852337386459112, 0.849511785898358, 0.00185880088247359, 0.0418722757603973, 0.736728538526222, 0.716277709929273, 0.903043946484104, 0.250322033185512, 0.97288141772151, 0.656528037041426, 0.25402183458209, 0.284065989777446, 0.729632006026804, 0.763572447933257, 0.0348138057161123, 0.00871321861632168, 0.43232419504784, 0.915044548455626, 0.438290276098996, 0.475907567422837, 0.439849420450628, 0.328594570048153, 0.0551980517338961, 0.481894522439688, 0.990172139601782, 0.654888461576775, 0.867879716679454, 0.343642072286457, 0.484055617591366, 0.220405105268583, 0.563844583695754, 0.76837615808472, 0.151773197343573, 0.585039120865986, 0.151443333365023, 0.992829578928649, 0.564717927481979, 0.848141403403133, 0.895290921442211, 0.966566636227071, 0.38202979369089, 0.289981779409572, 0.73494361480698, 0.063147830311209, 0.484161529922858, 0.53451741929166, 0.237481635063887, 0.603351338533685, 0.194983235094696, 0.000853131292387843, 0.4250520511996, 0.0773701993748546, 0.174938246840611, 0.730620980029926, 0.0433168599847704, 0.13176017999649, 0.532993183005601, 0.854394077556208, 0.261429066769779, 0.32215709425509, 0.984246356645599, 0.904214195674285, 0.59335330221802, 0.928315387340263, 0.926859020721167, 0.108630033675581, 0.668023525737226, 0.491975797805935, 0.604994338704273, 0.518783424980938, 0.0923375084530562, 0.95577231189236, 0.723749452037737, 0.388117723865435, 0.522513040574268, 0.774141357745975, 0.548468115273863, 0.677871701074764, 0.198412939440459, 0.627653139876202, 0.915030242875218, 0.250162169570103, 0.63649320951663, 0.630099521251395, 0.437956541543826, 0.260594762163237, 0.479751887498423, 0.0101860263384879, 0.868903384311125, 0.278534265235066, 0.0127006606198847, 0.905890410300344, 0.954413228202611, 0.777732153655961, 0.423686332534999, 0.319661873392761, 0.0363942580297589, 0.0628941715694964, 0.0980959604494274, 0.448384969960898, 0.870382471941411, 0.694613342639059, 0.98861387767829, 0.8272689813748, 0.548565677599981, 0.392956854775548, 0.016659040004015, 0.382538552395999, 0.908943108515814, 0.687715317122638, 0.754476502304897, 0.527856393251568, 0.52043259749189, 0.803923511179164, 0.523317163111642, 0.991189760621637, 0.282389135332778, 0.279642809880897, 0.968953379895538, 0.977746878750622, 0.894448178820312, 0.617621906334534, 0.336144367931411, 0.930004365276545, 0.245034691644832, 0.307021783431992, 0.33591337595135, 0.381661017425358, 0.616874752100557, 0.349720158847049, 0.864073160802945, 0.812214081175625, 0.773558815941215, 0.0332281226292253, 0.720803753472865, 0.333394939079881, 0.293997866800055, 0.210136831039563, 0.625551791163161, 0.125645183725283, 0.278940967982635, 0.19114456162788, 0.584939487278461, 0.309831542661414, 0.278195457300171, 0.291209308896214, 0.961860785260797, 0.562040177173913, 0.406785474624485, 0.548569195903838, 0.00591545412316918, 0.812543980311602, 0.561635501915589, 0.0461486109998077, 0.0913120692130178, 0.659805897623301, 0.463068171404302, 0.0019154807087034, 0.977813325589523, 0.638679464580491, 0.889903092756867, 0.667840568581596, 0.227132498752326, 0.59319547819905, 0.742457273649052, 0.939611955778673, 0.029191556153819, 0.072578297695145, 0.911583729553968, 0.564862226136029, 0.433506809873506, 0.23172370181419, 0.0492066368460655, 0.66766925691627, 0.632951707346365, 0.232738879742101, 0.0682831825688481, 0.647762133041397, 0.762648772913963, 0.311037149047479, 0.135462310863659, 0.101488806307316, 0.451145179569721, 0.440114471828565, 0.460363109828904, 0.504310304764658, 0.819494463736191, 0.368709030561149, 0.11292487103492, 0.49297067290172, 0.187894044909626, 0.846641366137192, 0.123257661238313, 0.0146676558069885, 0.83562449621968, 0.276052104542032, 0.749211929738522, 0.215082115493715, 0.731818155152723, 0.785989616066217, 0.510365455411375, 0.42040618439205, 0.912081928458065, 0.628477797610685, 0.0451976747717708, 0.88461459451355, 0.772413738071918, 0.292556818341836, 0.457560306414962, 0.282840635394678, 0.876900379080325, 0.490505006397143, 0.901587756117806, 0.0925958990119398, 0.373322947882116, 0.259347421815619, 0.99719837657176, 0.759156876942143, 0.589204312069342, 0.960889464244246, 0.570704887155443, 0.514403043780476, 0.370084909023717, 0.0324721976649016, 0.483091224916279, 0.0806573450099677, 0.25647241412662, 0.0174417593516409, 0.378499185666442, 0.115482739405707, 0.615778922801837, 0.982564938254654, 0.982259431155398, 0.694678434869274, 0.231022216146812, 0.15450671967119, 0.0717334633227438, 0.946030147140846, 0.0459343583788723, 0.104337275028229, 0.716064690379426, 0.505841404665262, 0.326971624279395, 0.554171376628801, 0.52933626784943, 0.357585409888998, 0.557354666292667, 0.98311444115825, 0.929871205240488, 0.498887814115733, 0.485765814315528, 0.507715021260083, 0.00228935573250055, 0.634705710224807, 0.0578073156066239, 0.346117743989453, 0.533338987268507, 0.918750796467066, 0.428956324700266, 0.452920676441863, 0.879929113434628, 0.3300540081691, 0.812715488485992, 0.169737463118508, 0.526181595399976, 0.239230300998315, 0.739062979118899, 0.0595781798474491, 0.865404638228938, 0.892479337984696, 0.360588231589645, 0.890929282875732, 0.0170817882753909, 0.18201761553064, 0.256273494334891, 0.171049508266151, 0.171611865982413, 0.527811765437946, 0.898018472827971, 0.425534424604848, 0.927834413014352, 0.00516632106155157, 0.987822852795944, 0.127195185050368, 0.910455838777125, 0.0126489924732596, 0.0701108567882329, 0.687154712388292, 0.793865685816854, 0.628855613525957, 0.783394184662029, 0.371452470310032, 0.162210805574432, 0.70195185369812, 0.262570054270327, 0.893886694218963, 0.09181077638641, 0.59755718242377, 0.599733222275972, 0.735746824881062, 0.547474856721237, 0.327091409591958, 0.783021450275555, 0.614899839274585, 0.285991258919239, 0.383444255916402, 0.814596732147038, 0.605047683930025, 0.324656182667241, 0.0563643714413047, 0.688327224226668, 0.220998305827379, 0.534116534050554, 0.0465474878437817, 0.658519071061164, 0.31795811210759, 0.0174901108257473, 0.269044719170779, 0.327655332861468, 0.552760693244636, 0.527267325436696, 0.931105934316292, 0.943559960229322, 0.136977499350905, 0.578164597041905, 0.97577673336491, 0.777235242072493, 0.0406325731892139, 0.902660964755341, 0.329938958166167, 0.325051829218864, 0.0868564648553729, 0.502117306226864, 0.67249808460474, 0.814258778933436, 0.491357359336689, 0.227366980398074, 0.889612680533901, 0.472413354320452, 0.104469970799983, 0.735417225863785, 0.230063383467495, 0.67311017611064, 0.769934271229431, 0.899370427709073, 0.697003066539764, 0.693082979181781, 0.873396060196683, 0.895496719516814, 0.576876376755536, 0.936046404996887, 0.612219240050763, 0.319894244428724, 0.298719848971814, 0.674479691311717, 0.324314448283985, 0.536971733206883, 0.834297828841954, 0.502639674348757, 0.197790765902027, 0.0318340570665896, 0.432790343184024, 0.917945949360728, 0.0568419666960835, 0.246917149052024, 0.840562190162018, 0.753172199008986, 0.821550698252395, 0.434009949909523, 0.29115403839387, 0.592664239229634, 0.203143415274099, 0.826835412764922, 0.100309215951711, 0.425427143927664, 0.635221774922684, 0.214761895593256, 0.142829260090366, 0.543172727804631, 0.281252494547516, 0.866224286844954, 0.597124559571967, 0.565247613238171, 0.687336225993931, 0.396031925687566, 0.00861812452785671, 0.174927197396755, 0.562591426074505, 0.0897084758616984, 0.705764994490892, 0.260751003399491, 0.250611269148067, 0.78215402062051, 0.562833839096129, 0.479941909667104, 0.807414822047576, 0.886707232799381, 0.208783611888066, 0.494374809786677, 0.0977038608398288, 0.250623839674518, 0.736020325450227, 0.324449447216466, 0.397117839660496, 0.70361799118109, 0.663066530134529, 0.103325493633747, 0.832683072192594, 0.579547961242497, 0.311789801809937, 0.738825378473848, 0.088018219685182, 0.468150246655568, 0.567103378474712, 0.546556472079828, 0.784525115275756, 0.388798929750919, 0.604504637885839, 0.789020592812449, 0.553451858926564, 0.194205911830068, 0.421866525895894, 0.0444178150501102, 0.480440902058035, 0.0219743249472231, 0.579217916587368, 0.80300835845992, 0.620315968757495, 0.405010092072189, 0.597791714128107, 0.887159829260781, 0.462306142086163, 0.309790004976094, 0.4655068856664, 0.0629158089868724, 0.713130056392401, 0.54938556230627, 0.184015497099608, 0.691340161254629, 0.681318345945328, 0.939093822380528, 0.744577007833868, 0.240134551888332, 0.207875113235787, 0.209614632185549, 0.00509716244414449, 0.197905265260488, 0.984625758603215, 0.316108469851315, 0.421823306009173, 0.602068784879521, 0.225741620408371, 0.412584307603538, 0.696386907948181, 0.827314416179433, 0.622086374787614, 0.961064010160044, 0.0816554336342961, 0.156390789197758, 0.774857524782419, 0.294898208230734, 0.324949565343559, 0.639676910126582, 0.604059686185792, 0.197976938914508, 0.832598559558392, 0.257077305112034, 0.958244926063344, 0.0865679376292974, 0.402608836302534, 0.703206202480942, 0.0292445335071534, 0.265755149070174, 0.373700361698866, 0.79223708063364, 0.58241264289245, 0.00987934833392501, 0.732811946189031, 0.946022017858922, 0.229251010110602, 0.875636084936559, 0.852573242271319, 0.846036876784638, 0.571414139121771, 0.491421904647723, 0.798576653469354, 0.05934331798926, 0.111886638915166, 0.815731351263821, 0.892812795005739, 0.519689315697178, 0.674582049949095, 0.118013493716717, 0.231669925386086, 0.429674009792507, 0.0990819181315601, 0.657109084539115, 0.618377856910229, 0.396150422748178, 0.169317359570414, 0.54362332355231, 0.48133436916396, 0.267543685156852, 0.829108721809462, 0.365018416428939, 0.11931029567495, 0.184947313508019, 0.964026031084359, 0.82201218418777, 0.29407242895104, 0.628139129141346, 0.467539868550375, 0.998416071059182, 0.904876622604206, 0.135211459128186, 0.280275842640549, 0.382950095226988, 0.222780587850139, 0.131339818704873, 0.88737001339905, 0.906743321102113, 0.58422586414963, 0.167149228043854, 0.932274113642052, 0.949846578296274, 0.927148988004774, 0.330940830986947, 0.0440037448424846, 0.371621413389221, 0.0627087843604386, 0.139742822153494, 0.677965085487813, 0.464029793860391, 0.916096061933786, 0.0599360645283014, 0.616870227269828, 0.190739104757085, 0.323916576337069, 0.934364258078858, 0.779072212288156, 0.733688439941034, 0.783321055117995, 0.241752208676189, 0.161356404423714, 0.594004932092503, 0.42478014761582, 0.46908904495649, 0.858485994162038, 0.204015175113454, 0.204347578808665, 0.00852191867306828, 0.208515252219513, 0.934749549720436, 0.396702929399908, 0.181994245620444, 0.819813223555684, 0.481858815532178, 0.755131945712492, 0.668255666969344, 0.901262130355462, 0.376325689256191, 0.146218504523858, 0.0058758354280144, 0.12886666925624, 0.794481817400083, 0.825380062917247, 0.406358007341623, 0.388821533881128, 0.268757042707875, 0.448067207820714, 0.309809993719682, 0.458947784733027, 0.885194010101259, 0.121640580240637, 0.994619208620861, 0.752838558284566, 0.576842007692903, 0.327774637611583, 0.118154135532677, 0.186412709532306, 0.580003298586234, 0.789081651484594, 0.737394368974492, 0.0948689044453204, 0.896513084648177, 0.747891929699108, 0.113424900453538, 0.604322905186564, 0.749036460882053, 0.0119681486394256, 0.884837853955105, 0.122698470717296, 0.575051304651424, 0.210526859620586, 0.445494061335921, 0.0573008763603866, 0.316703848075122, 0.592869674321264, 0.398251063190401, 0.368001338094473, 0.301786362193525, 0.572006300790235, 0.265708089806139, 0.580194022273645, 0.81531287333928, 0.486403964459896, 0.185954987071455, 0.183656794019043, 0.222917466424406, 0.403855997370556, 0.908013856969774, 0.995961501263082, 0.172174669103697, 0.921413158997893, 0.482720515457913, 0.892299243016168, 0.797945961821824, 0.659501477144659, 0.308821144979447, 0.365729942684993, 0.675168130779639, 0.989525465993211, 0.00493419845588505, 0.977133649168536, 0.899579596705735, 0.450945124262944, 0.00578402471728623, 0.0265030732844025, 0.599180517252535, 0.636229263618588, 0.395931607112288, 0.0894557097926736, 0.334311753278598, 0.163390068570152, 0.0176704102195799, 0.538746672449633, 0.532160534756258, 0.97687522880733, 0.0619266598951072, 0.32056373776868, 0.7475845660083, 0.533752959221601, 0.327971051679924, 0.172352060908452, 0.725402846699581, 0.426547006936744, 0.910944182658568, 0.895590811036527, 0.471301220357418, 0.096337276045233, 0.45159408589825, 0.993484654696658, 0.226369870593771, 0.168704340932891, 0.888468439690769, 0.126888601342216, 0.183711386518553, 0.481866441201419, 0.745115539757535, 0.532788030570373, 0.506366005400196, 0.448054821230471, 0.833007477689534, 0.374726709444076, 0.467123741749674, 0.316495993640274, 0.594729523407295, 0.435878166696057, 0.0844940727110952, 0.0148440494667739, 0.548075012862682, 0.320130300009623, 0.754781035007909, 0.519736085319892, 0.330443121260032, 0.226049386197701, 0.24137655692175, 0.643321553710848, 0.268905020551756, 0.233499272726476, 0.0233339285477996), actual = c(1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.844955608714372, 0.0833560775499791, 0.733897674828768, 0.338085582945496, 0.421938464744017, 0.617770605254918, 0.75372909149155, 0.638314454350621, 0.836275930749252, 0.90492036100477, 0.521664471831173, 0.425080332905054, 0.56834634160623, 0.937483730725944, 0.688481794903055, 0.895448438357562, 0.48234824417159, 0.0203531575389206, 0.905929853674024, 0.598937430651858, 0.629720795201138, 0.1275305734016, 0.00861564371734858, 0.443036758573726, 0.486037071794271, 0.131046328227967, 0.62786535709165, 0.586992608383298, 0.360811251215637, 0.735978022450581, 0.575800933409482, 0.964484441326931, 0.873124683042988, 0.64764792448841, 0.122419609455392, 0.11944637214765, 0.77909394283779, 0.423690674360842, 0.676463093608618, 0.00464764982461929, 0.900061825755984, 0.0332763167098165, 0.961079758359119, 0.148673096904531, 0.55972629156895, 0.1602700764779, 0.275597786530852, 0.834215359762311, 0.814645412843674, 0.492101444397122, 0.835755824809894, 0.230858408147469, 0.778424265561625, 0.290888300864026, 0.310033047106117, 0.405194590101019, 0.439093202818185, 0.840186877408996, 0.0684093201998621, 0.372755185235292, 0.228879816131666, 0.705583676928654, 0.686751822009683, 0.317023713840172, 0.933148602489382, 0.0350376358255744, 0.643030565464869, 0.0531456365715712, 0.563993241637945, 0.553744060220197, 0.55588989215903, 0.400569501100108, 0.516192645067349, 0.384501427179202, 0.827154844533652, 0.885968095855787, 0.234431903343648, 0.800686184316874, 0.329190424643457, 0.447400851873681, 0.909382925136015, 0.434932878008112, 0.879597691586241, 0.641900833463296, 0.489353240467608, 0.708136207191274, 0.403334320522845, 0.703508446924388, 0.32896297913976, 0.756719953613356, 0.110991129651666, 0.439842995721847, 0.883341434877366, 0.308716799598187, 0.993891931837425, 0.372131704818457, 0.63872266956605, 0.32683033356443, 0.95829686592333, 0.706933690933511, 0.822994099464267, 0.17019563773647, 0.922388288425282, 0.488067957339808, 0.844497368438169, 0.773248607292771, 0.539242153055966, 0.103999259881675, 0.343749538762495, 0.956880377139896, 0.912967626703903, 0.202317687915638, 0.527263354277238, 0.394498259294778, 0.546072260243818, 0.917819189606234, 0.518333349144086, 0.905498841078952, 0.791427283780649, 0.729585739085451, 0.680972623638809, 0.457810009596869, 0.386575725162402, 0.807198054855689, 0.0129385897889733, 0.629844149108976, 0.45761357084848, 0.640722758602351, 0.957224670099095, 0.613789920927957, 0.881352736381814, 0.338868203340098, 0.537794663337991, 0.69585557305254, 0.665260585723445, 0.333263904089108, 0.849531382555142, 0.618102860171348, 0.199505964294076, 0.197751396801323, 0.0617343063931912, 0.821549189276993, 0.496092278975993, 0.921972149983048, 0.878022135002539, 0.689208382274956, 0.0576304490678012, 0.996261480497196, 0.608458503847942, 0.396159983007237, 0.497031238628551, 0.540485270321369, 0.418876622105017, 0.343484997982159, 0.769237996079028, 0.943926075007766, 0.888910247012973, 0.292553967796266, 0.0116156558506191, 0.369838170474395, 0.139651985606179, 0.242336824769154, 0.301671328721568, 0.753269009059295, 0.719947340898216, 0.422278714599088, 0.474821951938793, 0.244590608403087, 0.825738870771602, 0.831477257655933, 0.400837156921625, 0.344357843976468, 0.194972933968529, 0.922902273014188, 0.596276141703129, 0.491825199918821, 0.163531743222848, 0.99870796315372, 0.357240454293787, 0.475978075992316, 0.710556623991579, 0.237274519400671, 0.116968015674502, 0.063874096609652, 0.792685320135206, 0.96055064513348, 0.901472165482119, 0.289338663453236, 0.457305221585557, 0.970912754302844, 0.227260357700288, 0.694541326258332, 0.717784368433058, 0.799807239789516, 0.00510311685502529, 0.025943684624508, 0.83228462934494, 0.968411919893697, 0.014626007527113, 0.10215133195743, 0.74033573246561, 0.753002676879987, 0.312291509006172, 0.107553724898025, 0.392874031327665, 0.351279823109508, 0.162484025815502, 0.407761706737801, 0.736115158768371, 0.304005928570405, 0.943779938155785, 0.461233861744404, 0.718951782910153, 0.056792642455548, 0.777758322190493, 0.950895139016211, 0.345309515250847, 0.626702786190435, 0.0625114156864583, 0.72528839064762, 0.24867546884343, 0.460959797957912, 0.361292326822877, 0.542556904023513, 0.401636556256562, 0.677721257787198, 0.895127943484113, 0.279517002403736, 0.646364144049585, 0.314283254323527, 0.871286522829905, 0.574220852227882, 0.0287475066725165, 0.808727033669129, 0.333930153399706, 0.262741061626002, 0.701851013349369, 0.0514720941428095, 0.0420444377232343, 0.359266734914854, 0.919448506319895, 0.784196730470285, 0.224754564231262, 0.292446068255231, 0.690643436741084, 0.394966998603195, 0.108233662787825, 0.844373975647613, 0.801837035221979, 0.988522605504841, 0.169779705582187, 0.398941455408931, 0.359684345545247, 0.599006332689896, 0.310752813937142, 0.587729990715161, 0.923607428558171, 0.446592785883695, 0.25305703910999, 0.27794925798662, 0.364918814273551, 0.149621117394418, 0.120906316209584, 0.503211550414562, 0.763340883655474, 0.571742211002856, 0.11610143003054, 0.806325372308493, 0.664883813122287, 0.189370250329375, 0.276722854934633, 0.689619857817888, 0.997143185697496, 0.630812798859552, 0.380969776539132, 0.56461791228503, 0.175925990566611, 0.302490106318146, 0.197375167859718, 0.594736848957837, 0.130844904342666, 0.367540986509994, 0.503265933832154, 0.441851150011644, 0.69899145164527, 0.707977042067796, 0.967291733482853, 0.0652237401809543, 0.209137408994138, 0.01646618405357, 0.212864564266056, 0.374515095725656, 0.276658461196348, 0.129652048461139, 0.683216557838023, 0.844793853582814, 0.933316333219409, 0.607845414662734, 0.582382657099515, 0.636316678253934, 0.682852430269122, 0.330857327906415, 0.661822573980317, 0.462431272724643, 0.966027684509754, 0.942179785110056, 0.0429634978063405, 0.761506827548146, 0.302696927450597, 0.330552969826385, 0.221490308409557, 0.360580228734761, 0.946272848872468, 0.0148087900597602, 0.240612020716071, 0.561463352525607, 0.495717110810801, 0.822038859594613, 0.141768168192357, 0.598935993388295, 0.734512655064464, 0.118546688463539, 0.454773663775995, 0.375601267442107, 0.721481185872108, 0.28541269688867, 0.630352126667276, 0.58379309787415, 0.696162159088999, 0.427842879202217, 0.440720619633794, 0.0304268752224743, 0.72433403506875, 0.157370882341638, 0.829226527363062, 0.349281018134207, 0.163068466819823, 0.46480222325772, 0.759670442668721, 0.16090535512194, 0.252746348036453, 0.896003864705563, 0.540884526679292, 0.835841137683019, 0.976495714858174, 0.285617249552161, 0.71791380061768, 0.584145459113643, 0.471496197627857, 0.0071457726880908, 0.354812340578064, 0.0562186064198613, 0.195902562700212, 0.21537180733867, 0.57647429802455, 0.414565312676132, 0.283611899707466, 0.260516617912799, 0.849667981499806, 0.7353774914518, 0.0652068462222815, 0.188785159727558, 0.741407789522782, 0.375619779108092, 0.926678121555597, 0.80055907741189, 0.562675085151568, 0.644685741746798, 0.346042467048392, 0.093717374606058, 0.0349075177218765, 0.280648624990135, 0.629409016110003, 0.30885784025304, 0.178135604364797, 0.763031130889431, 0.035447720438242, 0.548830128507689, 0.651143454946578, 0.991621775785461, 0.968610414303839, 0.412449280032888, 0.836541493190452, 0.719972187420353, 0.314217285951599, 0.682201253715903, 0.701282535912469, 0.123068868881091, 0.358969727065414, 0.25785566563718, 0.440724314888939, 0.554655368207023, 0.770921737654135, 0.959529778687283, 0.92554874625057, 0.0768900541588664, 0.974185130791739, 0.316958964569494, 0.590133215300739, 0.338539149146527, 0.682155333226547, 0.151633125729859, 0.538319922750816, 0.944612651364878, 0.359424435999244, 0.931303221033886, 0.313393241493031, 0.53340523969382, 0.681538038887084, 0.359343750402331, 0.343952554278076, 0.271859587635845, 0.785871308296919, 0.191316823707893, 0.233415785245597, 0.180553023004904, 0.157810247503221, 0.607700533466414, 0.63004999095574, 0.972754820948467, 0.609961340436712, 0.867445211857557, 0.14752377779223, 0.0977973300032318, 0.76970388344489, 0.0433408876415342, 0.154038906097412, 0.690542859956622, 0.529955620411783, 0.141918441280723, 0.225843629334122, 0.962052870541811, 0.449539767578244, 0.394538431428373, 0.647589200176299, 0.905376901384443, 0.383975816657767, 0.83613712945953, 0.341593745164573, 0.144576320890337, 0.668720531510189, 0.712148882448673, 0.4942355544772, 0.570013689342886, 0.358952883630991, 0.679783287923783, 0.240080159157515, 0.837103913305327, 0.713851546868682, 0.719265046063811, 0.46682886290364, 0.543305474566296, 0.97880738042295, 0.0510835282038897, 0.673295621294528, 0.107375402702019, 0.553245374001563, 0.258190958993509, 0.807766324607655, 0.228984169196337, 0.996010425267741, 0.102947807172313, 0.113203455461189, 0.707984167151153, 0.535269648768008, 0.715209822170436, 0.257571535184979, 0.571427480084822, 0.422300657490268, 0.590961516369134, 0.445364559534937, 0.0935429129749537, 0.16096956236288, 0.593512878986076, 0.621450055623427, 0.249240233097225, 0.92325569386594, 0.697607691399753, 0.327178670559078, 0.735171557869762, 0.904876562766731, 0.211794901639223, 0.547225665301085, 0.721966778859496, 0.846007835585624, 0.83440296840854, 0.912334325956181, 0.124712811317295, 0.388933630194515, 0.102747294353321, 0.169714738847688, 0.368882951559499, 0.551473147701472, 0.113135599065572, 0.803651077672839, 0.656283588847145, 0.502475933870301, 0.342734438832849, 0.977691319771111, 0.420117718400434, 0.445779291912913, 0.751069269143045, 0.52988130855374, 0.866947742411867, 0.595915379468352, 0.682827433571219, 0.348289453424513, 0.233089933404699, 0.548304841155186, 0.276530324947089, 0.103582441341132, 0.132644407916814, 0.506930452305824, 0.762146060355008, 0.866680365987122, 0.0105406027287245, 0.0958543978631496, 0.320790144847706, 0.612926606088877, 0.11070434213616, 0.0725257000885904, 0.0887864239048213, 0.637218414572999, 0.739441010635346, 0.986223538871855, 0.10629534185864, 0.814968535443768, 0.568449403392151, 0.167726203799248, 0.664713901467621, 0.387420958140865, 0.594432337209582, 0.586015243316069, 0.588024991797283, 0.28373829764314, 0.159923336934298, 0.539224676787853, 0.183882988989353, 0.298787329811603, 0.28525187401101, 0.648003410315141, 0.918340825708583, 0.44386089569889, 0.604224242502823, 0.981508773053065, 0.414346037898213, 0.5860182386823, 0.921272763283923, 0.102958135306835, 0.480383546091616, 0.382134274113923, 0.959216098533943, 0.594402904622257, 0.551371145062149, 0.471682552015409, 0.615931809181347, 0.97360328072682, 0.926760852336884, 0.668928259052336, 0.944136664038524, 0.796623202739283, 0.243100174935535, 0.519992602523416, 0.254995597992092, 0.353400875581428, 0.749361687339842, 0.0898779437411577, 0.747044450603426, 0.319484783569351, 0.274948228616267, 0.309078481281176, 0.0415931099560112, 0.293479479150847, 0.90824147243984, 0.201108382549137, 0.112330310279503, 0.208164087496698, 0.0914739156141877, 0.308724468806759, 0.305570245953277, 0.979317899560556, 0.893959838198498, 0.866154883056879, 0.673220367869362, 0.812355261761695, 0.819269403815269, 0.585019326768816, 0.526294134091586, 0.768808189546689, 0.229259412735701, 0.724581721704453, 0.819517317228019, 0.129597659688443, 0.0262902211397886, 0.442089447053149, 0.408038150984794, 0.629313593031839, 0.779132275143638, 0.739231413928792, 0.154417234938592, 0.628452997421846, 0.416579994373024, 0.30951758287847, 0.628006257349625, 0.42057302640751, 0.0744732853490859, 0.375704487087205, 0.461085093906149, 0.77836849889718, 0.953091752715409, 0.584032530663535, 0.16442947415635, 0.172860426595435, 0.159395046299323, 0.910446765366942, 0.619770832359791, 0.205359273124486, 0.115138251567259, 0.441293727606535, 0.521975784562528, 0.389686502050608, 0.27130400785245, 0.121957501629367, 0.158086626557633, 0.661163405748084, 0.245129975490272, 0.231801726855338, 0.0831933943554759, 0.748259477084503, 0.411480772541836, 0.710507625713944, 0.750333494972438, 0.536852873396128, 0.841619773302227, 0.155283727683127, 0.063942531356588, 0.0440867182333022, 0.203540140064433, 0.414589516818523, 0.103526602033526, 0.402587748831138, 0.213866696925834, 0.275795118883252, 0.362649526447058, 0.847797265974805, 0.824085960164666, 0.134251428302377, 0.432978028664365, 0.974517946364358, 0.968961549689993, 0.547724711941555, 0.851443774532527, 0.392343332991004, 0.938123415224254, 0.99383107596077, 0.187785194953904, 0.951810632133856, 0.768475798657164, 0.974078225903213, 0.91680502332747, 0.73225314100273, 0.795864577637985, 0.00317914318293333, 0.089584888657555, 0.776610726723447, 0.326056172605604, 0.950208717491478, 0.890089266933501, 0.501627791905776, 0.395065716002136, 0.0288730831816792, 0.244525021174923, 0.00375040457583964, 0.105466638458893, 0.482571618864313, 0.640044729225338, 0.314035739982501, 0.988384446827695, 0.287223995430395, 0.240286935120821, 0.396857840474695, 0.258312524529174, 0.688495001057163, 0.425557398470119, 0.472618622938171, 0.657582288607955, 0.568742668721825, 0.759612418944016, 0.772510906448588, 0.230152911506593, 0.732025489211082, 0.503990691155195, 0.720061026047915, 0.8882025629282, 0.720897728111595, 0.775981186656281, 0.75743808480911, 0.324706681771204, 0.220156942494214, 0.189242203021422, 0.267248570220545, 0.217698340537027, 0.176538497209549, 0.330281181028113, 0.655513985780999, 0.026725897565484, 0.20041302870959, 0.918429464334622, 0.513981722062454, 0.936457615578547, 0.859429504722357, 0.832137624500319, 0.576610439456999, 0.284243361558765, 0.278001059778035, 0.245217165211216, 0.946886187419295, 0.0114537291228771, 0.381167641840875, 0.905475483974442, 0.746222267858684, 0.424839514307678, 0.915158827789128, 0.266705485060811, 0.893027348443866, 0.897550932131708, 0.0705992749426514, 0.924958356423303, 0.967643328709528, 0.206930368673056, 0.056513785617426, 0.196938344277442, 0.807819734327495, 0.720702023711056, 0.375228641089052, 0.799422142794356, 0.264207555446774, 0.465850941371173, 0.782361519988626, 0.845628113718703, 0.71112775686197, 0.0578967675101012, 0.681147180264816, 0.952143598580733, 0.323245711624622, 0.814932413864881, 0.00515995617024601, 0.706355075584725, 0.957554230699316, 0.439826556481421, 0.402674145530909, 0.057102220132947, 0.367982986848801, 0.325189927825704, 0.254242516355589, 0.924069389235228, 0.481183772906661, 0.920645964564756, 0.810968041187152, 0.999922292307019, 0.348056671908125, 0.964944728417322, 0.285602138610557, 0.70031159510836, 0.443840152118355, 0.627220965689048, 0.225547035224736, 0.0814687723759562, 0.79176114150323, 0.211365095805377, 0.292762134457007, 0.89394280128181, 0.877687527565286, 0.276193879544735, 0.182025607209653, 0.89288875204511, 0.944724034750834, 0.329426062526181, 0.299332292750478, 0.0981663770508021, 0.147865245118737, 0.75026583718136, 0.835299172671512, 0.853690869174898, 0.984813985880464, 0.912921334849671, 0.127871468663216, 0.332234798930585, 0.0936394659802318, 0.88630448281765, 0.0956553730648011, 0.856670498847961, 0.159669219749048, 0.187678617425263, 0.722487918334082, 0.734380644978955, 0.470506426645443, 0.51880969456397, 0.418502227636054, 0.396494403015822, 0.323016120120883, 0.616821076022461, 0.0805748077109456, 0.11644656653516, 0.827329484047368, 0.461680588545278, 0.115195979364216, 0.665508563164622, 0.161759527167305, 0.530332386493683, 0.556238346500322, 0.285419185645878, 0.560418446548283, 0.842042893869802, 0.661862924462184, 0.393433099146932, 0.793901677010581, 0.488663312513381, 0.531332404352725, 0.953520404174924, 0.55363225704059, 0.8851472097449, 0.0821008589118719, 0.400899335043505, 0.760859698755667, 0.625703404657543, 0.969842524733394, 0.111852368805557, 0.338011833839118, 0.143395395250991, 0.975485527422279, 0.429925659438595, 0.982238893862814, 0.3911544282455, 0.356302493019029, 0.442411080701277, 0.68432019255124, 0.877694363472983, 0.872746103676036, 0.430363699793816, 0.0968337270896882, 0.118164754705504, 0.145539183169603, 0.611291954992339, 0.464241831563413, 0.545829481212422, 0.764214280759916, 0.161657066084445, 0.817955072270706, 0.17493770387955, 0.194742261199281, 0.0920156070496887, 0.814537858124822, 0.840126247378066, 0.675495469477028, 0.929551659617573, 0.800890840822831, 0.957345102448016, 0.991409438662231, 0.253576996736228, 0.155131732812151, 0.573208504356444, 0.624064102536067, 0.107283418765292, 0.44563146145083, 0.978499598801136, 0.682820919668302, 0.110901376698166, 0.731228297110647, 0.330180225195363, 0.218245431315154, 0.361211089184508, 0.214288912946358, 0.838256047572941, 0.53269354859367, 0.135924624279141, 0.66343659022823, 0.178192875348032, 0.740028335480019, 0.749873538501561, 0.386570900678635, 0.790002323687077, 0.561306527815759, 0.697246499825269, 0.211742696119472, 0.202910395339131, 0.304357416229323, 0.675051860744134, 0.0766923222690821, 0.677067777141929, 0.416338526876643, 0.95255191763863, 0.31264591217041, 0.00612612208351493, 0.00546118314377964, 0.400334178237244, 0.821420888416469, 0.208674660418183, 0.484732895391062, 0.676469631958753, 0.136141157243401, 0.0274355008732527, 0.829915302805603, 0.0293845504056662, 0.773693462600932, 0.988813762320206, 0.761807219358161, 0.632720647379756, 0.111926593352109, 0.428305132314563, 0.232057285960764, 0.845221837982535, 0.826806305907667, 0.399085319135338, 0.869319034274668, 0.603510842425749, 0.811872067395598, 0.32430088124238, 0.295341084711254, 0.983871174044907, 0.167454197304323, 0.898168507264927, 0.387110720388591, 0.95411273627542, 0.772018475225195, 0.249721494968981, 0.595106809167191, 0.99039466585964, 0.233279336476699, 0.0974584887735546, 0.158131275326014, 0.543870700290427, 0.352287161163986, 0.0010732221417129, 0.613295597722754, 0.0282952671404928, 0.58612299291417, 0.628985265037045, 0.629030432086438, 0.775863063521683, 0.164645178243518, 0.625679210294038, 0.86120091797784, 0.0809673878829926, 0.00766128627583385, 0.0658434322103858, 0.273425790015608, 0.218926722649485, 0.856020416133106, 0.475467304931954, 0.263992199674249, 0.550297207431868, 0.547863526036963, 0.621034855954349, 0.961161256534979, 0.962333976523951, 0.234443628462031, 0.508482853649184, 0.985425533959642, 0.120271364226937, 0.372915624873713, 0.780223835492507, 0.173037743195891, 0.730926441494375, 0.233825885923579, 0.312078108778223, 0.113984152907506, 0.906811537686735, 0.118464513681829, 0.969658092828467, 0.966544574126601, 0.0157712106592953, 0.882517060963437, 0.901883091544732, 0.202767205657437, 0.473076262511313, 0.244203846203163, 0.917705950094387, 0.703653219155967, 0.829371735453606, 0.605991941178218, 0.163579834857956, 0.654298214009032, 0.763231259072199, 0.728011323371902, 0.0924746287055314, 0.692402243381366, 0.272026104386896, 0.14989296416752, 0.222030091565102, 0.973245376488194, 0.0373091916553676, 0.766848385566846, 0.239556970307603, 0.20510702393949, 0.85370196797885, 0.917417620308697, 0.0646721264347434, 0.0304489047266543, 0.0903521021828055), actual = c(0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"