Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.7711137        0.4472755 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4718103        0.4718103        0.5221418 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.239156982395798, 0.396667403401807, 0.598858462646604, 0.67178447288461, 0.375827393960208, 0.256291196448728, 0.77212611050345, 0.83780892402865, 0.403470376972109, 0.811378567479551, 0.717935889028013, 0.575268165906891, 0.20021220529452, 0.262488051084802, 0.218264091294259, 0.365654096007347, 0.300599161768332, 0.0676607282366604, 0.895861884811893, 0.977900458965451, 0.369265361921862, 0.0538985354360193, 0.607704802881926, 0.507204532623291, 0.183453432982787, 0.51689595123753, 0.990097351837903, 0.441150246420875, 0.0252437375020236, 0.907304490217939, 0.0943070244975388, 0.480769441230223, 0.707962827989832, 0.247272961540148, 0.127271135337651, 0.789940222166479, 0.524653375148773, 0.84803482121788, 0.354927452513948, 0.809216605266556, 0.186072265030816, 0.99673346336931, 0.900736810173839, 0.0845853509381413, 0.596246489323676, 0.824038827558979, 0.571611711289734, 0.513160506496206, 0.25409313547425, 0.684951596194878, 0.790437521412969, 0.0519435007590801, 0.103076610714197, 0.0631057030986995, 0.574634323595092, 0.548507580300793, 0.488475564168766, 0.566088408697397, 0.211079799337313, 0.483028035378084, 0.687346847495064, 0.458685984835029, 0.660985748749226, 0.938216140493751, 0.329313910100609, 0.717462805565447, 0.999015169916674, 0.461262797471136, 0.498331816401333, 0.916537103010342, 0.400122852763161, 0.820918678771704, 0.56028154422529, 0.620151101145893, 0.994094974594191, 0.113517774501815, 0.234982942696661, 0.257435910170898, 0.610824723029509, 0.510241238167509, 0.508617171784863, 0.272623913828284, 0.683969331905246, 0.000717701157554984, 0.386162682436407, 0.893813697388396, 0.328305516857654, 0.529520522570238, 0.263752988073975, 0.386377015849575, 0.550732708768919, 0.205287972930819, 0.930484614335001, 0.865837243851274, 0.26908390596509, 0.566233805147931, 0.674296604702249, 0.775371686788276, 0.850118812639266, 0.439242871711031, 0.592290855012834, 0.251480631064624, 0.226328154327348, 0.298951570410281, 0.614437133539468, 0.126835893606767, 0.815616128500551, 0.11820709798485, 0.9233941975981, 0.856104078702629, 0.7080319870729, 0.504971930058673, 0.204309517517686, 0.608123157173395, 0.990397604415193, 0.452167446026579, 0.768649568082765, 0.345446308609098, 0.643831567838788, 0.0638978399801999, 0.0980312039609998, 0.28000591811724, 0.182818095432594, 0.891241917852312, 0.797988646663725, 0.0902634528465569, 0.057261505862698, 0.963566638994962, 0.495756206801161, 0.566645504906774, 0.0602443895768374, 0.58227615733631, 0.775609331205487, 0.451392134418711, 0.864505938719958, 0.86666799406521, 0.919417970813811, 0.491328697418794, 0.97901715664193, 0.484631076222286, 0.0585370177868754, 0.897053254768252, 0.421705526066944, 0.0361982411704957, 0.170503984903917, 0.619042232166976, 0.262392374221236, 0.483081734506413, 0.460245442111045, 0.569933380931616, 0.771043098531663, 0.605655098101124, 0.862913498654962, 0.209122523665428, 0.127033372409642, 0.484788675326854, 0.221947671612725, 0.816512216581032, 0.738124889088795, 0.989523450145498, 0.454962748335674, 0.117546053836122, 0.12294711265713, 0.459844291675836, 0.503694373881444, 0.85123255290091, 0.994946179445833, 0.393186392029747, 0.854958737269044, 0.824912367621437, 0.0515334189403802, 0.303519853390753, 0.790234145475551, 0.984425435308367, 0.541761325672269, 0.801453152438626, 0.202878184150904, 0.494483678368852, 0.134369509993121, 0.855811433400959, 0.48636123794131, 0.340542653342709, 0.343086784007028, 0.210181341506541, 0.352627475513145, 0.369465502444655, 0.450998058309779, 0.608387157786638, 0.129060019738972, 0.34483906859532, 0.518756902078167, 0.942932807607576, 0.730240868870169, 0.373808551346883, 0.848499461775646, 0.0685445822309703, 0.687031825305894, 0.563349148491398, 0.360028743511066, 0.243303088238463, 0.593824083218351, 0.504084447864443, 0.0709479881916195, 0.987426769454032, 0.731641531921923, 0.33522603707388, 0.711637974716723, 0.346148704178631, 0.663038227707148, 0.44200410717167, 0.0557457283139229, 0.593184812460095, 0.544452604372054, 0.104623976629227, 0.835429317085072, 0.839744545752183, 0.869669167324901, 0.462374930735677, 0.598846481414512, 0.471255913143978, 0.406837031943724, 0.821782204555348, 0.268987916409969, 0.401665229117498, 0.848133757710457, 0.458549577277154, 0.816847154404968, 0.54654232901521, 0.95303736301139, 0.185568216256797, 0.343747174600139, 0.541707913624123, 0.352535624057055, 0.43760303244926, 0.796581303002313, 0.580037728184834, 0.200572088826448, 0.0575499006081372, 0.937239221064374, 0.475526961963624, 0.857548546046019, 0.0438156917225569, 0.911085871746764, 0.937846993561834, 0.241243292111903, 0.889839109964669, 0.236110233236104, 0.762586321914569, 0.42678606742993, 0.668381892377511, 0.0476227204781026, 0.561369590694085, 0.575439520878717, 0.13960020407103, 0.47541699395515, 0.445214956300333, 0.859309827676043, 0.817340141162276, 0.0111539419740438, 0.29224340012297, 0.323165493551642, 0.726520305965096, 0.485351109411567, 0.168634447269142, 0.759688902180642, 0.0488774280529469, 0.584728642599657, 0.528186775045469, 0.608807219890878, 0.351219076197594, 0.0436136259231716, 0.971717922249809, 0.0142113477922976, 0.943153233034536, 0.684264921350405, 0.328742585144937, 0.966622828040272, 0.00494967517443001, 0.316191010409966, 0.570965806720778, 0.355730914976448, 0.719942447729409, 0.474392985692248, 0.0529248903039843, 0.835036961128935, 0.152057436062023, 0.86878653219901, 0.946738728322089, 0.00268293754197657, 0.537642325740308, 0.876343006500974, 0.547603302868083, 0.0749294485431165, 0.818111315602437, 0.839558746200055, 0.125817493768409, 0.380685675889254, 0.869212949648499, 0.979843199253082, 0.722489967476577, 0.341383858351037, 0.692600342677906, 0.810819356702268, 0.73514944780618, 0.336778200697154, 0.755376854911447, 0.988427196396515, 0.173040447058156, 0.600220384541899, 0.680869407951832, 0.573851243825629, 0.940065949689597, 0.168984124436975, 0.174714374821633, 0.743161336984485, 0.692195043200627, 0.2894558662083, 0.881075004581362, 0.12234002770856, 0.678213995881379, 0.196484811138362, 0.760884304065257, 0.721252304269001, 0.786096287891269, 0.670842784456909, 0.32702288380824, 0.99814568599686, 0.338514030445367, 0.599640634842217, 0.438360297586769, 0.855170674156398, 0.176356296055019, 0.590815737610683, 0.58640871103853, 0.255823093932122, 0.730155159020796, 0.339394338894635, 0.54104144196026, 0.783473379909992, 0.602574648102745, 0.885455571115017, 0.00766301923431456, 0.833133452571929, 0.0320696686394513, 0.504307916620746, 0.04854089114815, 0.471552094677463, 0.214748677331954, 0.4818983937148, 0.0361261409707367, 0.177475112723187, 0.423973428318277, 0.54875667556189, 0.851194593589753, 0.129715169314295, 0.672374103218317, 0.0518015653360635, 0.382074603112414, 0.176572434836999, 0.574737845454365, 0.179271915229037, 0.590569005347788, 0.362667207373306, 0.686630167998374, 0.329610922141001, 0.18326086178422, 0.326279630418867, 0.692029687575996, 0.439345853636041, 0.155195451341569, 0.638372539775446, 0.303462998475879, 0.603055747225881, 0.963686316041276, 0.498936874093488, 0.315078345593065, 0.45141000347212, 0.961343184811994, 0.278916634386405, 0.825580598786473, 0.216338750673458, 0.863218353362754, 0.982531879562885, 0.89492929331027, 0.996472621103749, 0.648195255780593, 0.906842914642766, 0.99546624510549, 0.0274880619253963, 0.723669913597405, 0.134916226379573, 0.670400557806715, 0.445514815393835, 0.18075834820047, 0.90957396151498, 0.275569338584319, 0.938944693654776, 0.331691021099687, 0.261295908596367, 0.389118035091087, 0.0719312008004636, 0.300524440594018, 0.802194978110492, 0.869205871131271, 0.448314138455316, 0.879745917161927, 0.942330217920244, 0.98317761789076, 0.904303622199222, 0.516127947950736, 0.407847085501999, 0.0263216341845691, 0.876834214664996, 0.299383325269446, 0.882360639283434, 0.183575168484822, 0.611816847231239, 0.589172440115362, 0.527399333659559, 0.543412180850282, 0.97747144731693, 0.497874912107363, 0.639627091120929, 0.782437233487144, 0.507394759450108, 0.547968016704544, 0.094779027858749, 0.430881587788463, 0.189533163793385, 0.755961849587038, 0.844210342038423, 0.196073339320719, 0.626619324553758, 0.558224141830578, 0.681474037934095, 0.956387909362093, 0.12013962934725, 0.89184286724776, 0.623449798906222, 0.8184951946605, 0.463949737139046, 0.393036084948108, 0.923276646295562, 0.635124973487109, 0.768966078292578, 0.0723369491752237, 0.494453935418278, 0.847054927609861, 0.711945814546198, 0.307936471654102, 0.167624728986993, 0.00105592468753457, 0.54635945847258, 0.015868044225499, 0.728164310567081, 0.12084085890092, 0.507707474753261, 0.581558017991483, 0.474448202876374, 0.913467740174383, 0.928300031693652, 0.43729345081374, 0.715566214406863, 0.227997865760699, 0.0326798723544925, 0.101216795155779, 0.728994106408209, 0.149846192449331, 0.224985794862732, 0.0156714005861431, 0.779557319357991, 0.0965538884047419, 0.572843107394874, 0.361720443470404, 0.903397452551872, 0.535428793402389, 0.900605809176341, 0.498625143663958, 0.61336573283188, 0.0253857548814267, 0.807191347237676, 0.891378559637815, 0.856517637148499, 0.784704832127318, 0.906951727811247, 0.415136199444532, 0.471162831410766, 0.110062200576067, 0.151636846596375, 0.900231015635654, 0.173613710794598, 0.271913767792284, 0.592850136803463, 0.884568495908752, 0.882682784227654, 0.00859619118273258, 0.785190557595342, 0.386180841363966, 0.90421473258175, 0.729461969109252, 0.512796994298697, 0.000961782643571496, 0.596623761812225, 0.519481781637296, 0.236507613444701, 0.107742946827784, 0.181098364060745, 0.00372865749523044, 0.630714193917811, 0.205094288336113, 0.464088668348268, 0.199701768578961, 0.78038334986195, 0.958362993551418, 0.0579213164746761, 0.259417661698535, 0.788074058946222, 0.745019752532244, 0.188900576438755, 0.97985426010564, 0.679118687752634, 0.531423227628693, 0.0499169963877648, 0.867429540725425, 0.621532930526882, 0.94770204462111, 0.427289178362116, 0.253401317866519, 0.873305808985606, 0.09919548407197, 0.700868338579312, 0.641717358957976, 0.466632127063349, 0.692528815474361, 0.129286521812901, 0.134642490884289, 0.666813936084509, 0.943935056449845, 0.281575926113874, 0.424697265960276, 0.599962988868356, 0.0641726197209209, 0.853719440056011, 0.702883710386232, 0.393030278384686, 0.241599716478959, 0.119933378649876, 0.963427414419129, 0.106333818519488, 0.585990495048463, 0.826429428067058, 0.571461187442765, 0.792275760089979, 0.466079793870449, 0.254669760586694, 0.925843341741711, 0.00189117388799787, 0.996009424794465, 0.253860982367769, 0.183062804862857, 0.296531138941646, 0.302995512960479, 0.732774575706571, 0.153135683853179, 0.326161755947396, 0.740824269829318, 0.880863724974915, 0.2157186656259, 0.303111516172066, 0.900978738674894, 0.873890607617795, 0.50466015515849, 0.914224951295182, 0.0263653511647135, 0.223327323561534, 0.294727480039001, 0.723891500849277, 0.502399979159236, 0.857841516612098, 0.921353459591046, 0.451652037212625, 0.802800305187702, 0.530944995814934, 0.826771901687607, 0.113426284398884, 0.63280751905404, 0.604077960364521, 0.50915544340387, 0.274611680302769, 0.792352359043434, 0.54765012091957, 0.0856751024257392, 0.916165037546307, 0.658751946873963, 0.737242548726499, 0.081128143472597, 0.609461655374616, 0.263874415308237, 0.713488837704062, 0.865535677876323, 0.160210493020713, 0.341363583225757, 0.485835231374949, 0.481095728464425, 0.787333725718781, 0.252231874736026, 0.342984877293929, 0.415731006069109, 0.0660191478673369, 0.956709692953154, 0.371350670000538, 0.840667449636385, 0.609172702999786, 0.817574797663838, 0.443054996430874, 0.601365312933922, 0.337282560067251, 0.314453067723662, 0.634738981723785, 0.844436548650265, 0.733261083019897, 0.366107751615345, 0.25947575783357, 0.676797834224999, 0.511613862356171, 0.201498431852087, 0.338050675345585, 0.481243028538302, 0.179843224352226, 0.15216153813526, 0.52300790627487, 0.294407097855583, 0.872563710901886, 0.588335333857685, 0.911603403044865, 0.614693215349689, 0.46287217666395, 0.41747585311532, 0.250826211180538, 0.916456871200353, 0.220683064544573, 0.38590265926905, 0.302910134661943, 0.600517179584131, 0.0825355427805334, 0.948168635368347, 0.330904116854072, 0.844470494892448, 0.712939799996093, 0.204311941750348, 0.702185778645799, 0.503917566500604, 0.464699185919017, 0.268471070332453, 0.852487126132473, 0.645912773907185, 0.0945764353964478, 0.447175153298303, 0.231464952463284, 0.15435777255334, 0.594815873773769, 0.429225269937888, 0.449088219087571, 0.77348065469414, 0.881567790871486, 0.127144898287952, 0.241916364524513, 0.557256676256657, 0.654545300872996, 0.728681049542502, 0.683345585828647, 0.717794650467113, 0.630089176353067, 0.553971509682015, 0.266443596221507, 0.382717094384134, 0.0318473898805678, 0.262942215427756, 0.110365088097751, 0.340591533575207, 0.813780194381252, 0.065577580826357, 0.276824540924281, 0.724451178219169, 0.453810859005898, 0.17233807221055, 0.367467272328213, 0.478783155791461, 0.502821725327522, 0.0326535399071872, 0.340248638996854, 0.326120540034026, 0.109309325227514, 0.250844311434776, 0.155043300241232, 0.915702618891373, 0.94385386724025, 0.165191767271608, 0.471989055164158, 0.218731707893312, 0.310600017663091, 0.89141991827637, 0.941043083323166, 0.964088019914925, 0.132551294518635, 0.645678804721683, 0.622872875537723, 0.56540100928396, 0.618059918517247, 0.669356401544064, 0.44028386566788, 0.887644192669541, 0.0549259735271335, 0.146978737320751, 0.124610569095239, 0.308221877552569, 0.854567473987117, 0.562579101417214, 0.268544599181041, 0.868551335530356, 0.70117268897593, 0.698550281813368, 0.0399263149593025, 0.169356492115185, 0.464978380827233, 0.106186563381925, 0.105440321611241, 0.974230799823999, 0.560954954707995, 0.880444415844977, 0.997327040415257, 0.281955242156982, 0.456291157053784, 0.968621080508456, 0.61700491909869, 0.109845706028864, 0.498723853845149, 0.556006822502241, 0.56987346848473, 0.180133104790002, 0.821845386177301, 0.750942644895986, 0.384867026237771, 0.102346323663369, 0.772456706967205, 0.632445212220773, 0.495421887142584, 0.914766648085788, 0.169987244298682, 0.860313745681196, 0.530487520620227, 0.995430702110752, 0.444966126233339, 0.841629171511158, 0.558340008370578, 0.469482148764655, 0.451560690067708, 0.151068014092743, 0.59771681111306, 0.129732799716294, 0.19876896077767, 0.892918005120009, 0.318422365700826, 0.141462556552142, 0.284737446345389, 0.744446703465655, 0.58755059610121, 0.862290978664532, 0.425242115277797, 0.392895421944559, 0.461645577801391, 0.0532220609020442, 0.52238020231016, 0.612231691135094, 0.322325752582401, 0.58177365642041, 0.383426282089204, 0.108845438575372, 0.503106582211331, 0.997492700116709, 0.525542145129293, 0.611762877553701, 0.336030754493549, 0.131201742449775, 0.816239986103028, 0.938079413492233, 0.0511478779371828, 0.776640088763088, 0.613888965453953, 0.305648511275649, 0.513933467911556, 0.87613919749856, 0.0132472445257008, 0.945599101250991, 0.108969478867948, 0.666212772019207, 0.130726518109441, 0.958190185949206, 0.0648828118573874, 0.536713306326419, 0.505952021339908, 0.140331858536229, 0.27746775909327, 0.620879139518365, 0.291738155065104, 0.26153594464995, 0.178799635730684, 0.603149808011949, 0.326258836081252, 0.439970582257956, 0.47043169522658, 0.478331614984199, 0.49283573566936, 0.893220765981823, 0.993066763039678, 0.783009666483849, 0.595292586833239, 0.10087171732448, 0.687999684363604, 0.645666918717325, 0.668413367355242, 0.844536387361586, 0.622542827855796, 0.332242600619793, 0.482711248798296, 0.669194829883054, 0.120938323205337, 0.849350688280538, 0.345640001352876, 0.50415317947045, 0.906431921990588, 0.722862773574889, 0.939483267487958, 0.636001974344254, 0.32298247423023, 0.269961501006037, 0.0881945143919438, 0.467911939602345, 0.025659236125648, 0.822754349559546, 0.768581694224849, 0.703949629561976, 0.933290096931159, 0.344679916277528, 0.825233939802274, 0.653641648823395, 0.242653286084533, 0.421784684760496, 0.948850424960256, 0.669265695149079, 0.19297037133947, 0.607899291440845, 0.775246871868148, 0.664477135287598, 0.361829583300278, 0.538099334109575, 0.372823406709358, 0.113087365170941, 0.133784476201981, 0.575266379863024, 0.269605071516708, 0.0295119504444301, 0.801625235471874, 0.721995397936553, 0.136822812259197, 0.627387683838606, 0.14593297848478, 0.509519674815238, 0.814173351740465, 0.868193821515888, 0.175411130767316, 0.712436978938058, 0.88382210675627, 0.893795361043885, 0.0794899584725499, 0.882228292757645, 0.532201101304963, 0.822429394815117, 0.0451085711829364, 0.238872488727793, 0.35899176215753, 0.23066916805692, 0.456930013373494, 0.0815452071838081, 0.709435680881143, 0.810076368041337, 0.310225866967812, 0.95615681563504, 0.630771335680038, 0.368093286873773, 0.701301993569359, 0.729738043155521, 0.77903852192685, 0.322377978125587, 0.480172640644014, 0.29309130157344, 0.278584021376446, 0.355013066902757, 0.195765759795904, 0.242477922467515, 0.212378055788577, 0.264764569466934, 0.0344716345425695, 0.97201051726006, 0.45148458192125, 0.861853000009432, 0.340677733300254, 0.979530944954604, 0.404529931023717, 0.746073498390615, 0.162979007931426, 0.154526274418458, 0.389217001618817, 0.540300421882421, 0.278765821829438, 0.142698175506666, 0.691560068866238, 0.217996790772304, 0.839169217273593, 0.960317577468231, 0.0346798547543585, 0.439394807675853, 0.41116181993857, 0.493781028082594, 0.490562567254528, 0.566813255427405, 0.553928722860292, 0.758891742210835, 0.854004105785862, 0.75765095395036, 0.664382396265864, 0.463492369279265, 0.883110435912386, 0.457392538199201, 0.776967339217663, 0.648622404783964, 0.510582768591121, 0.455750193214044, 0.717558485921472, 0.905040801037103, 0.071577196707949, 0.920081547927111, 0.19476752798073, 0.839744631201029, 0.129906370071694, 0.723552991403267, 0.112197435926646, 0.160737726837397, 0.886590290348977, 0.612020870205015, 0.265818056184798, 0.807880017440766, 0.186381590086967, 0.985086821485311, 0.738112138118595, 0.902443313738331, 0.750249265925959, 0.83312506461516, 0.14475065190345, 0.757807673187926, 0.253771414514631, 0.725063492776826, 0.0756447131279856, 0.808707450516522, 0.627802331000566, 0.0261211411561817, 0.344949882710353, 0.863712951773778, 0.131890836171806, 0.0269457586109638, 0.185377513524145, 0.917447672225535, 0.0832964119035751, 0.361987208481878, 0.987291514175013, 0.930510801961645, 0.959059571381658, 0.207272014347836, 0.471965859178454, 0.146249253535643, 0.443204897455871, 0.851313306484371, 0.240035453112796, 0.269347355002537, 0.0989579418674111, 0.656349306693301, 0.35730232601054, 0.239042634610087, 0.242603914812207, 0.869398314738646, 0.650209211744368, 0.456020673969761, 0.016122667118907, 0.888866851339117, 0.196753749158233, 0.909461191156879, 0.0144335920922458, 0.81722049228847, 0.935229799011722, 0.0154575887136161), actual = c(0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.285249282373115, 0.177065296564251, 0.358670692658052, 0.825811179354787, 0.0103270772378892, 0.81045689759776, 0.804065605858341, 0.489612980047241, 0.0802500797435641, 0.0910236423369497, 0.328287143958732, 0.325247885193676, 0.953043736051768, 0.175573189044371, 0.625977037707344, 0.920174854109064, 0.21854350506328, 0.363759152358398, 0.863925232551992, 0.773292544297874, 0.872965755173936, 0.984501195373014, 0.665543068666011, 0.906326572876424, 0.726648583542556, 0.0793146376963705, 0.309269331395626, 0.121527853887528, 0.267719876021147, 0.818129797931761, 0.459204263053834, 0.463169464608654, 0.989333592355251, 0.729421624680981, 0.410722591681406, 0.756870789686218, 0.110043145017698, 0.0976706475485116, 0.240436093183234, 0.117334761656821, 0.687779307132587, 0.697475883876905, 0.469847006490454, 0.650261613540351, 0.314069700893015, 0.148638541344553, 0.597992810187861, 0.0903877373784781, 0.74943958572112, 0.557665595551953, 0.724201420787722, 0.126763526815921, 0.100024091312662, 0.791727593634278, 0.923517561051995, 0.553968907566741, 0.612220689887181, 0.75061482610181, 0.712510437471792, 0.381314494647086, 0.566918689757586, 0.948377036722377, 0.790318529587239, 0.70831707585603, 0.545834590913728, 0.726259169867262, 0.961784420534968, 0.486822470556945, 0.0418179251719266, 0.409505432238802, 0.968719192547724, 0.803979923017323, 0.139512300724164, 0.743820488685742, 0.989187803119421, 0.375509748468176, 0.231974612455815, 0.143926332937554, 0.218966610264033, 0.966280624968931, 0.0791993644088507, 0.684565022122115, 0.94008245319128, 0.822558165760711, 0.262469676556066, 0.141909085446969, 0.359602337004617, 0.78445816738531, 0.537401508307084, 0.704470373922959, 0.940391889540479, 0.63631610549055, 0.23711215890944, 0.371211650082842, 0.973945579025894, 0.555261639179662, 0.396368255373091, 0.353053083410487, 0.578871031757444, 0.0732989124953747, 0.626471168827266, 0.241961205843836, 0.889303605305031, 0.737553890096024, 0.702054034220055, 0.494863625615835, 0.2654433993157, 0.790154186077416, 0.158942502690479, 0.898762817960232, 0.823658814188093, 0.955530691659078, 0.848038721596822, 0.714884958928451, 0.666646853322163, 0.359858426498249, 0.150769716128707, 0.959613934159279, 0.786128274630755, 0.742907298495993, 0.296935965074226, 0.808278865879402, 0.163536353036761, 0.281012063845992, 0.681261185090989, 0.329137721797451, 0.513535213423893, 0.452074186643586, 0.43313951510936, 0.188077107071877, 0.506522381212562, 0.619286638684571, 0.868219448020682, 0.0617667636834085, 0.595741203520447, 0.557584357913584, 0.4402208593674, 0.757973219035193, 0.727777779800817, 0.72122187516652, 0.883105856133625, 0.0796950561925769, 0.435394869884476, 0.0491776564158499, 0.344150785589591, 0.597755016991869, 0.726836060639471, 0.0485015420708805, 0.940231883665547, 0.924781687092036, 0.765118665993214, 0.823392397491261, 0.0491623526904732, 0.340062469709665, 0.285727218957618, 0.209281716961414, 0.4746547867544, 0.862762477248907, 0.561613593250513, 0.385839392663911, 0.530870873481035, 0.350769837619737, 0.341371836839244, 0.636611377121881, 0.690967777045444, 0.203924544388428, 0.403381715295836, 0.206422350136563, 0.748799366410822, 0.839045545784757, 0.247258944204077, 0.815371329430491, 0.245298754656687, 0.858145870268345, 0.148946733446792, 0.698643672047183, 0.766280700219795, 0.303863376611844, 0.240632075583562, 0.868762830505148, 0.383574071573094, 0.801980122923851, 0.234741677995771, 0.393224044237286, 0.887885093921795, 0.563895681640133, 0.253798588179052, 0.0335304930340499, 0.693886876339093, 0.469380719820037, 0.150754088535905, 0.343096880242229, 0.263645962113515, 0.294269521255046, 0.63526692870073, 0.783589697675779, 0.691819893429056, 0.698232119204476, 0.720785625977442, 0.648468045983464, 0.490651933010668, 0.49071423499845, 0.313985379645601, 0.0323688052594662, 0.0935203237459064, 0.987127462634817, 0.669552779057994, 0.595126856351271, 0.601103700464591, 0.667825469747186, 0.854520106222481, 0.187924526399001, 0.738700326532125, 0.922887437045574, 0.829379598377272, 0.518098737346008, 0.527525692246854, 0.58816903247498, 0.667240902082995, 0.501632256433368, 0.38453246303834, 0.343389759073034, 0.0111466178204864, 0.0177047303877771, 0.514664255781099, 0.407418189104646, 0.890325665939599, 0.333299984456971, 0.275651544565335, 0.0790722402743995, 0.255252926843241, 0.91000849683769, 0.908079748507589, 0.519415516406298, 0.320423886645585, 0.453357870224863, 0.915652940515429, 0.875892554409802, 0.867155356798321, 0.0244316819589585, 0.275304668350145, 0.156070972559974, 0.468600606778637, 0.42697179201059, 0.663236887659878, 0.698201055871323, 0.339734603418037, 0.734702009474859, 0.663105083396658, 0.215769510483369, 0.351235793204978, 0.675353314029053, 0.874648978002369, 0.227236761245877, 0.738387461751699, 0.0208705279510468, 0.987524170195684, 0.919506035046652, 0.02531678089872, 0.119121952913702, 0.902621155837551, 0.280561641557142, 0.200532373040915, 0.422276401892304, 0.163237493718043, 0.224023081595078, 0.14364339155145, 0.947016792371869, 0.971362615935504, 0.548802569275722, 0.452324679354206, 0.865838025230914, 0.721430921461433, 0.471601858967915, 0.0293262072373182, 0.732272936496884, 0.968415354145691, 0.739643670618534, 0.320034647127613, 0.16381905041635, 0.656876988708973, 0.358654419193044, 0.94161391002126, 0.829074357869104, 0.767134699737653, 0.188054407713935, 0.927995446836576, 0.34730404894799, 0.685298237716779, 0.412596564972773, 0.084000859875232, 0.244435485685244, 0.553601620020345, 0.25053449999541, 0.544140327954665, 0.28906082874164, 0.619486700044945, 0.330308215459809, 0.682281592627987, 0.114222459727898, 0.559189031133428, 0.777400519466028, 0.40125835314393, 0.80202806298621, 0.447311583440751, 0.944197520148009, 0.592657319502905, 0.61648602806963, 0.0581268270034343, 0.373394914204255, 0.001988610252738, 0.289492920041084, 0.337209888035432, 0.767361701698974, 0.0835022076498717, 0.158977133454755, 0.639635862549767, 0.3989729185123, 0.521921354578808, 0.116336781298742, 0.59118939936161, 0.362285957671702, 0.398242789087817, 0.369896387448534, 0.925963144516572, 0.438920773100108, 0.481886662775651, 0.317522086203098, 0.770514192525297, 0.973954395623878, 0.365516381105408, 0.485763327684253, 0.540735221467912, 0.572879880666733, 0.354239610023797, 0.0803385234903544, 0.661147288512439, 0.280354918679222, 0.68434537271969, 0.985335699981079, 0.966462097829208, 0.901637369766831, 0.917683301027864, 0.52256157877855, 0.620497454423457, 0.117666470585391, 0.462360336212441, 0.44368675397709, 0.355341046350077, 0.0743496934883296, 0.514454709831625, 0.456255614059046, 0.762449078727514, 0.709838877432048, 0.0338457620237023, 0.978706715628505, 0.148133108159527, 0.605348494136706, 0.428917936515063, 0.309808262856677, 0.557897130493075, 0.453722599195316, 0.665688045555726, 0.430784448981285, 0.863363356795162, 0.488575829425827, 0.704639327712357, 0.850710960337892, 0.510921923909336, 0.34551028162241, 0.144171292893589, 0.630509905517101, 0.336826007347554, 0.0480233684647828, 0.453179561067373, 0.945398269686848, 0.107366635464132, 0.807621479267254, 0.369451601523906, 0.38163576181978, 0.968104967148975, 0.724901434732601, 0.865046871127561, 0.228467582492158, 0.459129319759086, 0.733014379628003, 0.567641050089151, 0.158766514388844, 0.003871368477121, 0.333670472027734, 0.29113252251409, 0.776409515645355, 0.842403377173468, 0.47054181038402, 0.699501765659079, 0.026086444966495, 0.114006903488189, 0.309083085274324, 0.327642865013331, 0.44860279886052, 0.269378386670724, 0.575072256848216, 0.579234082018957, 0.0243238883558661, 0.607717981329188, 0.723430275684223, 0.112220999086276, 0.510041145374998, 0.304138832027093, 0.998034688876942, 0.744081832235679, 0.44516111863777, 0.0114373047836125, 0.469213217729703, 0.407417558366433, 0.343692400958389, 0.0843426662031561, 0.17015803954564, 0.407553960569203, 0.414888438070193, 0.498709672596306, 0.160147769376636, 0.55909777386114, 0.0368671321775764, 0.315605846699327, 0.535742298001423, 0.5668150188867, 0.104852468706667, 0.778429999249056, 0.267064433544874, 0.190679535269737, 0.757868856657296, 0.675132757984102, 0.482063050149009, 0.622138555860147, 0.189105309545994, 0.761185803450644, 0.95930018555373, 0.777016113977879, 0.839175740489736, 0.890972151653841, 0.498676970833912, 0.532415879424661, 0.465197142912075, 0.401982370065525, 0.345301176188514, 0.622300463262945, 0.25023432681337, 0.289904518751428, 0.957181619945914, 0.206317532574758, 0.524388145888224, 0.903316196519881, 0.763152002124116, 0.5737633719109, 0.493924151407555, 0.214491753606126, 0.134822483407333, 0.023272646125406, 0.182381694670767, 0.0554625780787319, 0.149427897064015, 0.586415625177324, 0.88458661432378, 0.663183018565178, 0.325539947487414, 0.319473148090765, 0.840670557226986, 0.160737627651542, 0.564315756550059, 0.670351265231147, 0.668993703555316, 0.105442113475874, 0.560438352404162, 0.674118808936328, 0.29784728307277, 0.297259460901842, 0.930730177788064, 0.195507952710614, 0.790255777537823, 0.10552168241702, 0.804557644762099, 0.420759382192045, 0.4570585899055, 0.521707942709327, 0.266792943235487, 0.0971155266743153, 0.665489838691428, 0.425879034446552, 0.8103089381475, 0.852494425373152, 0.93859742116183, 0.257475932594389, 0.131516783498228, 0.537562609650195, 0.0745083501096815, 0.987713498994708, 0.0942233349196613, 0.479263318935409, 0.300649106735364, 0.241376766236499, 0.664571418892592, 0.280923041282222, 0.918393242871389, 0.162255944916978, 0.276779961772263, 0.817747541237622, 0.368833559332415, 0.124892817577347, 0.304120200686157, 0.93938248232007, 0.442456418648362, 0.552962569287047, 0.350592303089797, 0.535335416439921, 0.664089924423024, 0.728729059686884, 0.374976122984663, 0.699917820747942, 0.308482000138611, 0.836531203007326, 0.508202879456803, 0.0717079683672637, 0.10085013625212, 0.050595753826201, 0.233328754780814, 0.531788932625204, 0.127403932157904, 0.525823806179687, 0.0820668919477612, 0.682693546637893, 0.3749877915252, 0.902250505518168, 0.686441169353202, 0.0861406077165157, 0.794347748393193, 0.355893431697041, 0.399105818942189, 0.470652007730678, 0.304716580780223, 0.726119469618425, 0.450249309651554, 0.810619469266385, 0.0229261904023588, 0.974695534911007, 0.776713395956904, 0.101543368306011, 0.285089217592031, 0.847945409128442, 0.116925017209724, 0.905285179615021, 0.521034163888544, 0.471149540273473, 0.490856892429292, 0.346090210136026, 0.400224237935618, 0.389810858760029, 0.206406589364633, 0.051085997140035, 0.83341611828655, 0.347268318058923, 0.507471184013411, 0.193767806747928, 0.0505782768595964, 0.0387377454899251, 0.234315593028441, 0.00563461799174547, 0.259709988022223, 0.56587350834161, 0.713094759266824, 0.608304605120793, 0.804936342872679, 0.517330770613626, 0.452995415078476, 0.170592323411256, 0.320289670489728, 0.292057890444994, 0.0828798513393849, 0.273224910022691, 0.984454650897533, 0.102562571875751, 0.210768891265616, 0.97971645090729, 0.271680084522814, 0.974836884299293, 0.207209450891241, 0.7720807637088, 0.767428830033168, 0.77846513222903, 0.675345688359812, 0.981895458651707, 0.494889810914174, 0.585286336485296, 0.151927490485832, 0.483339819358662, 0.151285968720913, 0.853055150713772, 0.820348609238863, 0.645080175483599, 0.532467867713422, 0.607898445567116, 0.891842825571075, 0.607576242648065, 0.789857350755483, 0.995409548515454, 0.832688969327137, 0.326274269493297, 0.949635158991441, 0.586066825548187, 0.905447592260316, 0.0937044224701822, 0.596449687844142, 0.09263258962892, 0.352655794704333, 0.449058691971004, 0.32289998489432, 0.0520412193145603, 0.190134491072968, 0.52865283889696, 0.749236574862152, 0.274183720350266, 0.0700286438222975, 0.916046463418752, 0.684852787293494, 0.230121845845133, 0.58369081094861, 0.852503336267546, 0.044160179561004, 0.147025066427886, 0.020023614866659, 0.99873602273874, 0.0025676020886749, 0.143448878778145, 0.932071974733844, 0.726051706587896, 0.828470821026713, 0.142686308594421, 0.298802773235366, 0.143105161376297, 0.903341767610982, 0.67447260976769, 0.422193127684295, 0.510668281232938, 0.960527097340673, 0.549991794861853, 0.285762469517067, 0.991337228566408, 0.787008256651461, 0.831493518780917, 0.807731297332793, 0.276461012195796, 0.731545319082215, 0.369047662708908, 0.771714065689594, 0.722458617063239, 0.0130097959190607, 0.395113923354074, 0.00264727394096553, 0.18885238096118, 0.84974160999991, 0.316304901847616, 0.667047954397276, 0.324290597811341, 0.496708982624114, 0.435531439492479, 0.346033165929839, 0.618467316031456, 0.0677664540708065, 0.711429990362376, 0.817205164348707, 0.117839281447232, 0.801089757820591, 0.487844661111012, 0.216855362057686, 0.401822978863493, 0.752167063532397, 0.612925636116415, 0.976214440539479, 0.341548313386738, 0.472751439083368, 0.389776772586629, 0.828485561069101, 0.603432254632935, 0.0911327954381704, 0.291881756857038, 0.576538054738194, 0.250624541193247, 0.674268936738372, 0.947508122306317, 0.561339381383732, 0.333182437112555, 0.28551904973574, 0.730278981151059, 0.711838815594092, 0.819436068180948, 0.395732732955366, 0.779990862123668, 0.253246289677918, 0.241276466753334, 0.726804648293182, 0.602919328492135, 0.496139732189476, 0.744295081589371, 0.216801311122254, 0.616245667217299, 0.269779400201514, 0.032017835881561, 0.155149951111525, 0.561926354886964, 0.493360250955448, 0.149158159038052, 0.335948234191164, 0.900672537041828, 0.580521747702733, 0.870385716669261, 0.964794639730826, 0.0835261617321521, 0.0267089232802391, 0.194819598691538, 0.585572025505826, 0.501819917233661, 0.738182868808508, 0.868355106795207, 0.820987034356222, 0.883750525303185, 0.820903122192249, 0.151018394622952, 0.624325590906665, 0.0262600753922015, 0.441242122091353, 0.43254787218757, 0.793843053979799, 0.635188628220931, 0.60803325031884, 0.127603070344776, 0.907659003976732, 0.734427570365369, 0.899901981465518, 0.650655468925834, 0.254273847676814, 0.155538027174771, 0.35961986426264, 0.723586157662794, 0.272670054575428, 0.129773067776114, 0.658325208351016, 0.651306602172554, 0.140412617707625, 0.600707744946703, 0.306707030627877, 0.000748910009860992, 0.696224752813578, 0.970878676278517, 0.974237746791914, 0.897283824859187, 0.0933122686110437, 0.343309497926384, 0.173373299650848, 0.410645656054839, 0.995727746747434, 0.400912820827216, 0.582689038710669, 0.874817304313183, 0.204482508823276, 0.381676686229184, 0.646916022757068, 0.193876998964697, 0.225541038904339, 0.995315574575216, 0.144292725715786, 0.93019643239677, 0.34612577711232, 0.490509341005236, 0.466237354325131, 0.240317831048742, 0.560074207140133, 0.826144861057401, 0.761091502849013, 0.265035579446703, 0.58611656492576, 0.648781809955835, 0.311738213989884, 0.633190519409254, 0.0939772329293191, 0.931434472324327, 0.905730162980035, 0.472099050879478, 0.74268106603995, 0.619004224194214, 0.827535330550745, 0.177853168454021, 0.863094129599631, 0.0293938932009041, 0.559841351350769, 0.283669487107545, 0.954631520900875, 0.74629432009533, 0.629753615008667, 0.20691279321909, 0.939379868330434, 0.524495555087924, 0.327950090402737, 0.430710199056193, 0.585155550856143, 0.00297692930325866, 0.260866221738979, 0.921852075494826, 0.0356184903066605, 0.218670093687251, 0.878824489191175, 0.0725612663663924, 0.632531737210229, 0.566386735532433, 0.0362546369433403, 0.738924899837002, 0.185773468343541, 0.129432681482285, 0.181058457819745, 0.693509433884174, 0.190332417842001, 0.321616419125348, 0.697856503073126, 0.270719401538372, 0.203984027728438, 0.532890893984586, 0.11775909177959, 0.487007658462971, 0.0420668832957745, 0.969582044752315, 0.566748176235706, 0.421667110174894, 0.412001689895988, 0.215982997789979, 0.17449085530825, 0.709846551297233, 0.67728747241199, 0.0206048521213233, 0.848441837821156, 0.490031878696755, 0.464787958655506, 0.229931069072336, 0.166466232854873, 0.233537259278819, 0.78900430793874, 0.0925764911808074, 0.102283017942682, 0.203268764540553, 0.924775908235461, 0.579555673524737, 0.0244454997591674, 0.944409953430295, 0.104414383880794, 0.76752347429283, 0.616485190344974, 0.81908225803636, 0.57605804479681, 0.918197825318202, 0.782786900177598, 0.685581922531128, 0.967670964775607, 0.382609696593136, 0.815047381911427, 0.790620666695759, 0.967451928416267, 0.0760016094427556, 0.554117998108268, 0.409236313775182, 0.3510445819702, 0.758210356812924, 0.997985705733299, 0.842649588128552, 0.515507803065702, 0.0765671250410378, 0.173460628138855, 0.532104519894347, 0.622497418895364, 0.443001972511411, 0.391793902264908, 0.529475459828973, 0.787135964725167, 0.0670792739838362, 0.808142768684775, 0.412285733502358, 0.183752269949764, 0.0274485263507813, 0.265492245554924, 0.313796745846048, 0.409460204420611, 0.871584526728839, 0.903766035102308, 0.667266291333362, 0.85123074403964, 0.420612855814397, 0.954317424912006, 0.323656165041029, 0.389717847341672, 0.982453634263948, 0.0133485018741339, 0.765040467958897, 0.426202536094934, 0.247106167254969, 0.285465185763314, 0.477650764165446, 0.594264247454703, 0.863836576463655, 0.914733027806506, 0.490155017236248, 0.803645299049094, 0.292116640368477, 0.70894984761253, 0.0716443515848368, 0.389165141619742, 0.742350533371791, 0.55474833631888, 0.806062743766233, 0.246332675917074, 0.027999913552776, 0.519571478711441, 0.437250298913568, 0.750528947683051, 0.960782576352358, 0.0312584557104856, 0.572315047960728, 0.176166338846087, 0.433745152549818, 0.0752351174596697, 0.389089471194893, 0.577668187906966, 0.914125091396272, 0.1153247973416, 0.382757345214486, 0.574653380317613, 0.891848949482664, 0.952867116080597, 0.694588632322848, 0.018938438501209, 0.0911878026090562, 0.888439155183733, 0.983726676087826, 0.958653625100851, 0.801207918208092, 0.861203197389841, 0.869509108131751, 0.721446769777685, 0.884051200933754, 0.874926935648546, 0.528735720319673, 0.724177976604551, 0.47022873791866, 0.343873341102153, 0.48041868256405, 0.799854957964271, 0.330112926196307, 0.32669303775765, 0.999346117256209, 0.00553232687525451, 0.0194297698326409, 0.261991132050753, 0.973031273111701, 0.87071605026722, 0.129510966362432, 0.686989062698558, 0.354693108703941, 0.693492879392579, 0.433842186117545, 0.577863656217232, 0.719012559158728, 0.88095245952718, 0.0217090484220535, 0.555347729474306, 0.0231302129104733, 0.686939631123096, 0.898946679197252, 0.133316811406985, 0.226287376368418, 0.13259958056733, 0.0371985849924386, 0.793239362537861, 0.171932694269344, 0.0505293749738485, 0.954844236141071, 0.2121866280213, 0.140811078017578, 0.735799996647984, 0.992078760638833, 0.53623402537778, 0.203036135295406, 0.0287109694909304, 0.432497355621308, 0.29240742046386, 0.0363684913609177, 0.841068172827363, 0.138995486311615, 0.592536964686587, 0.142421548487619, 0.760826451936737, 0.285531535511836, 0.14583410997875, 0.107217469019815, 0.999200568534434), actual = c(1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"