Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.8099239        0.8063918 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5031910        0.5031910        0.5031910 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.0623092430178076, 0.617084463359788, 0.884803770808503, 0.905638533178717, 0.578466058708727, 0.640090772882104, 0.879722363082692, 0.266188417794183, 0.497760070022196, 0.597421586280689, 0.596858019009233, 0.104600981110707, 0.532535589998588, 0.00161773897707462, 0.588660453446209, 0.450252066599205, 0.386073120171204, 0.249339074362069, 0.279458523960784, 0.154383454937488, 0.450997452950105, 0.410554593894631, 0.638367359293625, 0.29712355369702, 0.0508296331390738, 0.489208352286369, 0.757106702076271, 0.920928511535749, 0.762272882508114, 0.536865443224087, 0.78891544858925, 0.857956238789484, 0.850666769314557, 0.211730018025264, 0.320594993187115, 0.669956547673792, 0.450472751865163, 0.291003558551893, 0.660637671593577, 0.400783169548959, 0.179086746880785, 0.408763787476346, 0.381671564653516, 0.130295992130414, 0.539852248737589, 0.815899304579943, 0.813565041404217, 0.50607732241042, 0.602208384079859, 0.401302284328267, 0.24059397331439, 0.0881091023329645, 0.792830957099795, 0.433172242483124, 0.957091191783547, 0.665266555268317, 0.896258548600599, 0.739621794549748, 0.586097704945132, 0.647300184704363, 0.17911094496958, 0.639926324598491, 0.852674614172429, 0.191937803057954, 0.892334970412776, 0.179968087933958, 0.0223766174167395, 0.335602566134185, 0.316621909150854, 0.148843858391047, 0.0642185481265187, 0.349264466203749, 0.169240168761462, 0.902338894782588, 0.333641993347555, 0.479999938979745, 0.706657617352903, 0.540606907103211, 0.369067483814433, 0.583129320992157, 0.806059206370264, 0.359032477485016, 0.549913844326511, 0.283570440253243, 0.672646869672462, 0.091926989145577, 0.211395454825833, 0.342862147372216, 0.539244808256626, 0.989866485819221, 0.187969819409773, 0.475139153655618, 0.979698238195851, 0.0384573766496032, 0.907628184882924, 0.997412443161011, 0.346082689706236, 0.680285940179601, 0.344582257326692, 0.952634597895667, 0.0942845966201276, 0.888401229167357, 0.229827704606578, 0.85167795047164, 0.0632676384411752, 0.421688702655956, 0.872237208299339, 0.505479593295604, 0.247369839111343, 0.316995383473113, 0.328336140839383, 0.656967790797353, 0.24263742682524, 0.992119994014502, 0.742955935187638, 0.522599771153182, 0.556670094374567, 0.979466742370278, 0.45654755202122, 0.9474602045957, 0.224141323938966, 0.306752155767754, 0.0583877896424383, 0.432492467807606, 0.954387607751414, 0.520205473760143, 0.560637727379799, 0.227434324100614, 0.199956840835512, 0.263976567424834, 0.871781259775162, 0.402826659148559, 0.439254863187671, 0.0341554572805762, 0.00654744566418231, 0.284261601045728, 0.941780955763534, 0.798183388775215, 0.0817035106010735, 0.20424969564192, 0.191308955894783, 0.443178854882717, 0.341362602775916, 0.703557695960626, 0.383444934617728, 0.897796120727435, 0.369194108061492, 0.869926826562732, 0.121195176616311, 0.929102077847347, 0.988138951361179, 0.0671797837130725, 0.366083653410897, 0.477097101742402, 0.134051852161065, 0.347744538681582, 0.772869173670188, 0.724997723009437, 0.990845381049439, 0.0623689645435661, 0.857468623900786, 0.545343427220359, 0.817261993419379, 0.557796155568212, 0.123353477101773, 0.0789589053019881, 0.660030235303566, 0.743499987525865, 0.629022537497804, 0.336875081760809, 0.235922987340018, 0.546999466139823, 0.757784004788846, 0.0270936926826835, 0.31056928448379, 0.925174487754703, 0.770105318864807, 0.867967952042818, 0.935143342940137, 0.921910074539483, 0.957945535192266, 0.868432816583663, 0.310159062966704, 0.253834326285869, 0.855254770023748, 0.557790203951299, 0.723756264429539, 0.29088204796426, 0.488935436820611, 0.81654849066399, 0.269469621125609, 0.600872750394046, 0.616164952050894, 0.260266700526699, 0.540799516253173, 0.56545633636415, 0.272273307433352, 0.775459340307862, 0.641920840367675, 0.683425520313904, 0.81730094505474, 0.623872818658128, 0.537619530688971, 0.914108191849664, 0.139566261321306, 0.133816649904475, 0.326233019120991, 0.65852972259745, 0.0929864579811692, 0.379171001259238, 0.574706960236654, 0.273149556713179, 0.491348354145885, 0.155556039186195, 0.487814482999966, 0.979368552099913, 0.659324810374528, 0.852947314968333, 0.679495793068781, 0.0391614362597466, 0.588548729661852, 0.966897716280073, 0.267967401538044, 0.0455815668683499, 0.104766102042049, 0.405145616969094, 0.306684806244448, 0.339163053780794, 0.777618383057415, 0.994084922829643, 0.567314793821424, 0.0806733998470008, 0.1247215163894, 0.078842684160918, 0.39526390354149, 0.0468607293441892, 0.891192087903619, 0.896438221214339, 0.877475195331499, 0.796602019807324, 0.775751823792234, 0.744024233426899, 0.696472204988822, 0.246684986632317, 0.434373711235821, 0.559383400948718, 0.408777769654989, 0.237027229275554, 0.0248545575886965, 0.139961880631745, 0.65267722774297, 0.555062526138499, 0.229993045330048, 0.74153998401016, 0.845321608241647, 0.219206800218672, 0.725604834035039, 0.920417417539284, 0.783478380646557, 0.701200888259336, 0.554661920759827, 0.117423153482378, 0.0930823350790888, 0.461791650392115, 0.726807977305725, 0.670239487895742, 0.215953542618081, 0.653177216183394, 0.590807370608672, 0.880850036162883, 0.537642709678039, 0.146400356665254, 0.882376088295132, 0.870141715509817, 0.820276639424264, 0.222102792235091, 0.0198477364610881, 0.732182000763714, 0.337876002304256, 0.314826580230147, 0.602527468232438, 0.449987734202296, 0.0578659365419298, 0.402219901094213, 0.415333542041481, 0.871828661067411, 0.207254921086133, 0.763145393459126, 0.26659998181276, 0.903805942507461, 0.175416120095178, 0.0977613541763276, 0.391634655883536, 0.0586125787813216, 0.234126378549263, 0.808181597385556, 0.987056503305212, 0.297576313139871, 0.257424822775647, 0.330599296838045, 0.182901824591681, 0.902636644430459, 0.692572402767837, 0.371865293011069, 0.85481355083175, 0.457401786465198, 0.460761662805453, 0.725247056223452, 0.402326416922733, 0.949269405333325, 0.920259934617206, 0.316439936868846, 0.202179003506899, 0.32315035816282, 0.746953837107867, 0.776320294244215, 0.133097448386252, 0.846074004191905, 0.499649944715202, 0.866574702318758, 0.135506705148146, 0.752399222925305, 0.701103765284643, 0.319152778014541, 0.952827182365581, 0.092865003971383, 0.31555745867081, 0.0877479070331901, 0.685724272858351, 0.313481349498034, 0.706514071207494, 0.69558068132028, 0.15419593360275, 0.573584785684943, 0.976322837173939, 0.475333116250113, 0.338626619661227, 0.804174977354705, 0.600561931496486, 0.711534335510805, 0.555036119418219, 0.815762296784669, 0.667667559348047, 0.645610142033547, 0.517706280108541, 0.495696352561936, 0.794106935383752, 0.275623906869441, 0.428122446639463, 0.700982567621395, 0.0209438207093626, 0.743726255139336, 0.951668673427776, 0.554466853616759, 0.394974979804829, 0.375391955254599, 0.465063204290345, 0.748781511560082, 0.571866371668875, 0.0336843715049326, 0.332282879855484, 0.37840199097991, 0.675877980655059, 0.872306043049321, 0.136596331605688, 0.465384697541595, 0.803704395657405, 0.937465299386531, 0.904755762312561, 0.904364943737164, 0.1098328307271, 0.266188414068893, 0.919212304055691, 0.252054041484371, 0.668337020091712, 0.321621069684625, 0.112023257184774, 0.514939305372536, 0.633531713392586, 0.473843838786706, 0.71192962070927, 0.0559158260002732, 0.775688888737932, 0.101496103452519, 0.0311986326705664, 0.827344183577225, 0.12316710758023, 0.849603911628947, 0.353878860129043, 0.0348705905489624, 0.609767579939216, 0.586510899476707, 0.493191640591249, 0.801836457336321, 0.430244858609512, 0.710380900884047, 0.499104869784787, 0.0991428582929075, 0.459369051968679, 0.208010430680588, 0.326899179257452, 0.629246707540005, 0.918554857373238, 0.685124062001705, 0.887418768601492, 0.463657729560509, 0.89235313795507, 0.437321126228198, 0.0357228198554367, 0.928567362483591, 0.692238220013678, 0.0558302844874561, 0.0591372766066343, 0.101845865370706, 0.764358780812472, 0.650253099389374, 0.314567086752504, 0.838665734278038, 0.842366755474359, 0.0347078207414597, 0.91728591080755, 0.86320101423189, 0.694511178648099, 0.954389697406441, 0.617022081511095, 0.111635353881866, 0.764375962782651, 0.183427797630429, 0.108773099724203, 0.771189314778894, 0.786829400109127, 0.249588648788631, 0.830245484830812, 0.505602231482044, 0.108446032740176, 0.782602552557364, 0.845706333871931, 0.0647012800909579, 0.896307942224666, 0.661061552818865, 0.578288740711287, 0.689568155677989, 0.492606699699536, 0.11653699236922, 0.0325943916104734, 0.0497432823758572, 0.448342155199498, 0.191734810126945, 0.938066889764741, 0.390673023415729, 0.687496425816789, 0.870831156615168, 0.577728008385748, 0.997526767430827, 0.258216549176723, 0.813718242105097, 0.431723931571469, 0.240675654262304, 0.564466601470485, 0.962602107552812, 0.379449892090634, 0.155751917744055, 0.377968551591039, 0.834656846243888, 0.538756019435823, 0.641432270873338, 0.955853973748162, 0.906091531505808, 0.514738375786692, 0.918467355892062, 0.82661737408489, 0.756072726100683, 0.248977396870032, 0.347668730653822, 0.692871271399781, 0.285971341887489, 0.34456044412218, 0.473817702615634, 0.278121519368142, 0.63799003418535, 0.131115819560364, 0.86544405692257, 0.102963900892064, 0.0849303347058594, 0.742340207798406, 0.228190570604056, 0.887260071001947, 0.932995539391413, 0.849471034249291, 0.865090918028727, 0.529124551219866, 0.632970140315592, 0.534572633681819, 0.887249464634806, 0.86041394690983, 0.494295802665874, 0.742163971764967, 0.695164550794289, 0.52862645406276, 0.283692916389555, 0.273680926533416, 0.667158424155787, 0.293560053687543, 0.751687585376203, 0.551252843346447, 0.598170286277309, 0.644079928286374, 0.637749236775562, 0.329425096511841, 0.748970124172047, 0.815293561434373, 0.102036258438602, 0.283554689725861, 0.895472003845498, 0.358289879513904, 0.253442535875365, 0.216662393882871, 0.965421689441428, 0.845158628420904, 0.999695690115914, 0.450714853592217, 0.430419242242351, 0.13814385025762, 0.575986298499629, 0.7174099215772, 0.0474723514635116, 0.171036701416597, 0.577179320622236, 0.845620321109891, 0.487691930029541, 0.10618991823867, 0.495985009009019, 0.833460701163858, 0.749018486822024, 0.421514820074663, 0.780180405359715, 0.300776841817424, 0.301408663857728, 0.714852828998119, 0.336416317848489, 0.0087442877702415, 0.699646133230999, 0.100741807138547, 0.807190868305042, 0.410223391605541, 0.044546575518325, 0.156143934931606, 0.643332032253966, 0.59354325546883, 0.829037104733288, 0.479365831939504, 0.381921610794961, 0.780363767640665, 0.859852131921798, 0.382480673724785, 0.840243890881538, 0.310228482820094, 0.24854397540912, 0.8230181557592, 0.199064420070499, 0.819780713878572, 0.241884771035984, 0.66718534566462, 0.924430171493441, 0.0172916697338223, 0.119468145770952, 0.0914611076004803, 0.881464765872806, 0.17999638337642, 0.543265441898257, 0.680709460983053, 0.0197189026512206, 0.149894321803004, 0.255500250030309, 0.864106004592031, 0.969762517604977, 0.450421875109896, 0.707051416393369, 0.164655968779698, 0.733643685467541, 0.326680191792548, 0.481252337805927, 0.263531572418287, 0.569145564222708, 0.300893248291686, 0.0143873649649322, 0.427948919590563, 0.258268681354821, 0.100829533301294, 0.690276981098577, 0.949926390312612, 0.378526392625645, 0.0991730715613812, 0.762528151040897, 0.19394555571489, 0.48672264162451, 0.936348784947768, 0.690850184764713, 0.0312461119610816, 0.824630742194131, 0.0208811336196959, 0.297301919199526, 0.864661315223202, 0.956419061170891, 0.317536124493927, 0.766034255502746, 0.49592424556613, 0.168219317216426, 0.650321126682684, 0.481348977657035, 0.919942560372874, 0.834300328977406, 0.701307888375595, 0.685180423548445, 0.805923770181835, 0.0015477635897696, 0.10578660806641, 0.514567527920008, 0.0744257506448776, 0.769186333054677, 0.388543265638873, 0.218501657713205, 0.307137614814565, 0.7739891060628, 0.731692565139383, 0.25603836379014, 0.933245647931471, 0.137434587813914, 0.704019591910765, 0.769070929381996, 0.626729155657813, 0.991091448348016, 0.490357134025544, 0.743487534346059, 0.778007697314024, 0.318102277815342, 0.132205932633951, 0.429620149312541, 0.785061800153926, 0.54201672365889, 0.0645445163827389, 0.653149903286248, 0.415400776080787, 0.317679269704968, 0.545815393561497, 0.14128136751242, 0.13351716985926, 0.987878035521135, 0.872126508038491, 0.356042241677642, 0.534654723014683, 0.896454311208799, 0.0218308467883617, 0.75851269159466, 0.190625273389742, 0.804380708839744, 0.809521463233978, 0.854151965817437, 0.654000992886722, 0.244978495873511, 0.653368864906952, 0.592985604191199, 0.227360769640654, 0.900459759170189, 0.897067110519856, 0.891113444464281, 0.6126436998602, 0.606047480599955, 0.691028445959091, 0.186545953387395, 0.36305641871877, 0.0585505727212876, 0.14604459120892, 0.872934623854235, 0.293599932454526, 0.00691978237591684, 0.60366148664616, 0.650051314616576, 0.790550595382228, 0.803186032455415, 0.325184491463006, 0.0688087546732277, 0.357900810893625, 0.0914400776382536, 0.118592121871188, 0.158714968478307, 0.717742107575759, 0.970532868755981, 0.189214111771435, 0.755237623816356, 0.408166284905747, 0.904860629234463, 0.516745276516303, 0.582845660159364, 0.845881692599505, 0.81528262491338, 0.075893459841609, 0.0181920491158962, 0.491874478990212, 0.296445231186226, 0.263342685764655, 0.927886529825628, 0.970075592398643, 0.898756327340379, 0.218742748256773, 0.895730922929943, 0.840687130810693, 0.35085008176975, 0.94587242603302, 0.826825037132949, 0.189781996887177, 0.679685976123437, 0.0713399918749928, 0.392901067389175, 0.0856430425774306, 0.20500343455933, 0.606279807165265, 0.0284016653895378, 0.603575886460021, 0.998606637585908, 0.496442108880728, 0.0821325199212879, 0.874646567972377, 0.563439932186157, 0.0125261207576841, 0.87688622623682, 0.103518958901986, 0.769298871746287, 0.697111017070711, 0.348858988611028, 0.131237688707188, 0.813296406064183, 0.0254835856612772, 0.698621524032205, 0.408999133622274, 0.961119031067938, 0.373572233365849, 0.369471978861839, 0.550669000949711, 0.154625408817083, 0.532380205811933, 0.134312876034528, 0.848786926129833, 0.534574671415612, 0.592978912172839, 0.686910652788356, 0.377228737110272, 0.354467558674514, 0.940549897495657, 0.0216035286430269, 0.701958436751738, 0.970759770832956, 0.833353055873886, 0.87911710026674, 0.833802747074515, 0.732051406987011, 0.547853325726464, 0.0628433199599385, 0.16033204831183, 0.740630963351578, 0.0507702017202973, 0.00248685851693153, 0.188307280652225, 0.548231028486043, 0.310644639423117, 0.716583634959534, 0.617209436371922, 0.437721503898501, 0.506926117697731, 0.818437703652307, 0.432530564721674, 0.869765737326816, 0.981173648033291, 0.514053127029911, 0.392145839286968, 0.800594497937709, 0.0957368719391525, 0.474763731937855, 0.800412290496752, 0.889673445140943, 0.794410402653739, 0.629305740119889, 0.590751945273951, 0.726783753605559, 0.686042038956657, 0.103615418775007, 0.758596145780757, 0.651745219016448, 0.783004872966558, 0.337618229677901, 0.653372639790177, 0.955644887406379, 0.379721553064883, 0.25639320909977, 0.477448876947165, 0.343028635717928, 0.273440890945494, 0.588230226188898, 0.979475313797593, 0.949971198802814, 0.90607854234986, 0.0447267477866262, 0.472139218589291, 0.412686830153689, 0.926800278248265, 0.563197680516168, 0.707990920403972, 0.916238185018301, 0.683841774705797, 0.936853679595515, 0.637741392944008, 0.850172889884561, 0.787912276340649, 0.108665060717613, 0.558182824170217, 0.395383844617754, 0.726980873849243, 0.648845380404964, 0.0963974322658032, 0.878191788448021, 0.971144952112809, 0.660197890596464, 0.58223470277153, 0.529385612579063, 0.352269843220711, 0.0166052507702261, 0.0551856365054846, 0.624140023021027, 0.833405388286337, 0.910443195607513, 0.097153443377465, 0.0851097584236413, 0.669507859740406, 0.595198025694117, 0.37585722701624, 0.0346349587198347, 0.661603049375117, 0.844214051961899, 0.594378408510238, 0.118743875762448, 0.782426645047963, 0.748268079943955, 0.445010645315051, 0.485942487372085, 0.043262192979455, 0.456113856518641, 0.932229698169976, 0.745922474889085, 0.558048992184922, 0.392408765852451, 0.0753691368736327, 0.240130654769018, 0.93683788436465, 0.995770967798308, 0.3289115189109, 0.603050844743848, 0.760967866051942, 0.7192004728131, 0.885248031932861, 0.81282165274024, 0.327855400973931, 0.31664111209102, 0.79566903039813, 0.467394086765125, 0.632502161664888, 0.0580583759583533, 0.0413190240506083, 0.658451341558248, 0.762655598111451, 0.289084536256269, 0.110087913228199, 0.467949625570327, 0.675557573093101, 0.211416478035972, 0.527560736285523, 0.307287045056, 0.25151024851948, 0.47486529010348, 0.845408227527514, 0.862999110948294, 0.719335902947932, 0.139893799088895, 0.012834228342399, 0.0348716594744474, 0.445142110344023, 0.48489407915622, 0.64534642547369, 0.566268087131903, 0.170462721260265, 0.84399403934367, 0.386045533465222, 0.545744814444333, 0.0605946690775454, 0.745393407763913, 0.858678702730685, 0.811715955613181, 0.90332629927434, 0.138667927822098, 0.628097759094089, 0.241002339869738, 0.386153187835589, 0.431142937624827, 0.272006057202816, 0.168477279832587, 0.445734209846705, 0.0867089775856584, 0.972638686420396, 0.363726837793365, 0.970190517604351, 0.956501948880032, 0.450966314878315, 0.664043871453032, 0.109136950690299, 0.161139034898952, 0.804869413608685, 0.103522086748853, 0.796096282778308, 0.43030175101012, 0.81452324683778, 0.351545181823894, 0.794418281875551, 0.85791923222132, 0.842844805214554, 0.0265430356375873, 0.547611506422982, 0.0611505629494786, 0.923584925476462, 0.570473542204127, 0.0145154930651188, 0.853603613097221, 0.254376557888463, 0.0859650035854429, 0.148460250813514, 0.205763999605551, 0.323501880513504, 0.219963940326124, 0.616094038588926, 0.691857391037047, 0.272133531980217, 0.538729343097657, 0.403715048683807, 0.384355294052511, 0.350942895747721, 0.299588372698054, 0.165418643504381, 0.817102562403306, 0.345184046542272, 0.206062633311376, 0.945768871344626, 0.277276048902422, 0.474382813321427, 0.65793534158729, 0.745550952851772, 0.119064904050902, 0.92441392946057, 0.883617524988949, 0.221383576514199, 0.776890289969742, 0.957863666350022, 0.686540581285954, 0.935152041958645, 0.134439265588298, 0.797646227059886, 0.576303594280034, 0.054805257357657, 0.153722201474011, 0.58181191701442, 0.719046784797683, 0.220458663534373, 0.539523016428575, 0.965216469718143, 0.79370998730883, 0.937562379054725, 0.0147476918064058, 0.741230330429971, 0.471457472071052, 0.0570026615168899, 0.47777225333266, 0.732637653360143, 0.291403355076909, 0.640364982420579, 0.712457049870864, 0.485119619872421, 0.647126860450953, 0.322788922581822, 0.822019636165351, 0.0594012658111751, 0.50002633430995, 0.0868306958582252, 0.965932046994567, 0.920397744746879, 0.577486514346674, 0.916294985683635, 0.669075176119804, 0.22301541082561, 0.0166019354946911, 0.18359006382525, 0.296563747804612, 0.85560259828344, 0.868325205286965), actual = c(1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.420439090812579, 0.285404638154432, 0.369061195524409, 0.692209125263616, 0.610366056207567, 0.375318995909765, 0.176742712734267, 0.237294111866504, 0.683194954646751, 0.936590456403792, 0.848269171547145, 0.600855841767043, 0.904360787477344, 0.574126963037997, 0.124044023454189, 0.518178568920121, 0.297894304385409, 0.105611115461215, 0.0474398303776979, 0.438957509119064, 0.623261597007513, 0.566381750628352, 0.38714875979349, 0.15210624365136, 0.134599733632058, 0.307392918970436, 0.876913792686537, 0.274533319519833, 0.348058552946895, 0.366139103891328, 0.828280815854669, 0.61503521935083, 0.342176089296117, 0.721780218183994, 0.987650968832895, 0.249007662991062, 0.766493300674483, 0.286794033832848, 0.530567572219297, 0.198623805306852, 0.227820180822164, 0.564995527965948, 0.593026243383065, 0.293981182156131, 0.988496230216697, 0.630280325422063, 0.714143321150914, 0.817902718205005, 0.401146371848881, 0.747329051606357, 0.690582843963057, 0.904432555893436, 0.567675018450245, 0.627023031702265, 0.496426583267748, 0.388860718812793, 0.400698852492496, 0.66334483306855, 0.476752399466932, 0.10111402370967, 0.0919474281836301, 0.82509252615273, 0.600669662468135, 0.43615193432197, 0.332771147368476, 0.00806695502251387, 0.230189167428762, 0.133907608687878, 0.231883758446202, 0.547797365114093, 0.190521684242412, 0.394592192256823, 0.141264942474663, 0.302030398510396, 0.410979556385428, 0.145212086383253, 0.0514503761660308, 0.202331973705441, 0.489972949493676, 0.117793471319601, 0.888707812875509, 0.946636140579358, 0.112322804983705, 0.712915321346372, 0.319462177809328, 0.175199238350615, 0.135827552061528, 0.586946012685075, 0.929950444493443, 0.188400439685211, 0.324081420432776, 0.365813251817599, 0.602576743345708, 0.407729592639953, 0.565163139952347, 0.703745805425569, 0.995762401958928, 0.967452118173242, 0.607277854112908, 0.287371915299445, 0.653046091319993, 0.613818295067176, 0.375120911514387, 0.80955955456011, 0.599342523142695, 0.769081411417574, 0.426193918799981, 0.207962807035074, 0.738700177054852, 0.0968846923206002, 0.154868405777961, 0.657049672212452, 0.704010079614818, 0.982127330265939, 0.789115814724937, 0.44350756588392, 0.931686304975301, 0.566479417961091, 0.020390349207446, 0.845794074004516, 0.101938961772248, 0.750359107041731, 0.00431309919804335, 0.678898488404229, 0.177130446536466, 0.302874590968713, 0.870964834233746, 0.270294496323913, 0.168761001201347, 0.395388913573697, 0.587354570627213, 0.847911368124187, 0.0123146253172308, 0.646307992050424, 0.686610967852175, 0.730379056185484, 0.34838427323848, 0.111910683335736, 0.433461028384045, 0.0590984425507486, 0.154161349870265, 0.444057480664924, 0.897988044889644, 0.0593881774693727, 0.789469870273024, 0.455532925901935, 0.405927047599107, 0.566380989039317, 0.105712707387283, 0.184690835652873, 0.459290876053274, 0.146012484794483, 0.118079003877938, 0.295481089735404, 0.754034048179165, 0.660285961348563, 0.822991913883016, 0.663282339228317, 0.843536738539115, 0.137686802074313, 0.986494373995811, 0.666750488337129, 0.940788535866886, 0.779224436730146, 0.0043904900085181, 0.562254298478365, 0.334496389143169, 0.895196315366775, 0.372652116930112, 0.765020463382825, 0.257175958482549, 0.334612288046628, 0.156138060148805, 0.764730853028595, 0.851450585294515, 0.230192805640399, 0.0177256709430367, 0.926166447345167, 0.292264008196071, 0.313815707806498, 0.115821409504861, 0.529914127197117, 0.318432996515185, 0.872731224168092, 0.654478336917236, 0.414765898603946, 0.381748407380655, 0.919880580157042, 0.140525971306488, 0.157544408459216, 0.850958549883217, 0.230422974564135, 0.682362501742318, 0.739867915865034, 0.0936124671716243, 0.513409500475973, 0.543236467288807, 0.328492438187823, 0.525361921871081, 0.0337238244246691, 0.666196138830855, 0.389195661991835, 0.617346975952387, 0.499767773784697, 0.316685869824141, 0.433958175126463, 0.910249642794952, 0.888934668852016, 0.591172684216872, 0.240157606778666, 0.0727348148357123, 0.578251771163195, 0.457435561344028, 0.676713329041377, 0.900668733054772, 0.483734413282946, 0.680363749619573, 0.465507626300678, 0.23173603694886, 0.508453015005216, 0.400496192742139, 0.576399850659072, 0.694385490147397, 0.786631901515648, 0.816159784561023, 0.981136300601065, 0.0584200955927372, 0.112821265356615, 0.810808522859588, 0.764306384371594, 0.794150663539767, 0.335766006959602, 0.645630350802094, 0.128299917792901, 0.000699977623298764, 0.922485239105299, 0.167423055507243, 0.381503871642053, 0.818627127679065, 0.911179498303682, 0.833546534413472, 0.129622321343049, 0.162540993886068, 0.679969620192423, 0.267556566512212, 0.684013237245381, 0.0527498724404722, 0.633208290906623, 0.97113679186441, 0.220849974779412, 0.713008559076115, 0.0314348405227065, 0.921522634103894, 0.124639734160155, 0.72934726299718, 0.564881000900641, 0.100517785176635, 0.429552641930059, 0.103407789953053, 0.356693297857419, 0.705041937995702, 0.118798044743016, 0.667835821863264, 0.0618287562392652, 0.0583097613416612, 0.343251385260373, 0.375153610715643, 0.0175982546061277, 0.439793739002198, 0.266460559330881, 0.8714014289435, 0.332698369864374, 0.656323918839917, 0.061409876216203, 0.704659684095532, 0.594552671769634, 0.00131078599952161, 0.185688727069646, 0.663306177360937, 0.990045238519087, 0.234438349260017, 0.934099317993969, 0.578127299202606, 0.846936834976077, 0.0182899713981897, 0.859902189811692, 0.054244197672233, 0.750480476766825, 0.739822685951367, 0.207388771697879, 0.116466861683875, 0.331949949031696, 0.555969740729779, 0.795916521223262, 0.269387639360502, 0.719067866215482, 0.184330269694328, 0.843120745383203, 0.411592321004719, 0.109772766940296, 0.0657512398902327, 0.975305903470144, 0.919689271366224, 0.128425264032558, 0.586767105618492, 0.791734724771231, 0.467294221511111, 0.329297387273982, 0.901982703479007, 0.00980388093739748, 0.575572427362204, 0.911641741869971, 0.609097963664681, 0.899318699259311, 0.420485396869481, 0.659081066260114, 0.952153930440545, 0.152232822496444, 0.449732465669513, 0.611454759724438, 0.917855050414801, 0.77250380627811, 0.210856524994597, 0.916280725505203, 0.147728685289621, 0.237093509640545, 0.509405275806785, 0.702432482270524, 0.995373873738572, 0.969225756125525, 0.113475077785552, 0.266778342891484, 0.830983248306438, 0.789788040565327, 0.110818938817829, 0.947796132415533, 0.46846420224756, 0.102407564641908, 0.925125077366829, 0.458081058925018, 0.669593805680051, 0.504419186618179, 0.119359774049371, 0.516528881387785, 0.0622418390121311, 0.20518516888842, 0.325597224757075, 0.655295659555122, 0.265354066388682, 0.441718481481075, 0.823549597058445, 0.617824517190456, 0.231771798571572, 0.715365560492501, 0.316124340053648, 0.333761835005134, 0.719277809606865, 0.597326332936063, 0.0213116565719247, 0.995933897560462, 0.754789900733158, 0.0270317383110523, 0.676747547462583, 0.943112042732537, 0.385127632413059, 0.995121260406449, 0.267527645453811, 0.595619201892987, 0.156349606346339, 0.505417952081189, 0.965304275974631, 0.301068265922368, 0.0377593995071948, 0.900310860946774, 0.176573255797848, 0.858637136640027, 0.463878707494587, 0.724134732270613, 0.118590918136761, 0.135112970601767, 0.662350909318775, 0.580714773852378, 0.395967699587345, 0.482518942561001, 0.542415139498189, 0.682947233086452, 0.593279503751546, 0.0885563914198428, 0.195369510212913, 0.971246289787814, 0.251116731204093, 0.00372805027291179, 0.0915431813336909, 0.651382662821561, 0.642191098537296, 0.110357900848612, 0.153358194977045, 0.538462923374027, 0.997295506065711, 0.410481824073941, 0.507744750007987, 0.827734180027619, 0.994710799306631, 0.8470015716739, 0.420758683467284, 0.00507814832963049, 0.690450431546196, 0.581120503833517, 0.384783553658053, 0.552521522622555, 0.903539424529299, 0.371936626033857, 0.206682033604011, 0.824556306470186, 0.07430310244672, 0.755916721187532, 0.346008840948343, 0.803497302345932, 0.520603187149391, 0.741645767353475, 0.185268408851698, 0.168623733567074, 0.440246894489974, 0.193336986936629, 0.957952148746699, 0.907394979847595, 0.422976559959352, 0.784384557278827, 0.670982729876414, 0.665044780122116, 0.398009004769847, 0.467923507792875, 0.274631205480546, 0.150262816576287, 0.253712495556101, 0.50697468733415, 0.95366658619605, 0.358351787086576, 0.993155522737652, 0.481475684093311, 0.95861615636386, 0.543303288519382, 0.945863309781998, 0.679507811553776, 0.0438538657035679, 0.759357730392367, 0.876828740583733, 0.745778104057536, 0.751986628398299, 0.512342452770099, 0.0244536213576794, 0.702741236425936, 0.332071378827095, 0.994626969797537, 0.392448134953156, 0.590301233110949, 0.14088364620693, 0.238701976137236, 0.166617986746132, 0.0407106126658618, 0.703430151101202, 0.960203252499923, 0.425927201285958, 0.817721834173426, 0.406094497069716, 0.957581855589524, 0.773158120224252, 0.127836552448571, 0.605136079713702, 0.698717944091186, 0.899213519180194, 0.348950735060498, 0.292332894634455, 0.890307016903535, 0.5314099076204, 0.0495382894296199, 0.999656402971596, 0.206871367990971, 0.776956780813634, 0.0704163510818034, 0.298275106586516, 0.915623183595017, 0.53075785166584, 0.0027818561065942, 0.306616905843839, 0.656649672659114, 0.183481807587668, 0.419342272682115, 0.750164959114045, 0.79097414156422, 0.00696139293722808, 0.61090821470134, 0.879430335015059, 0.00704541127197444, 0.962311379378662, 0.623363734455779, 0.827487635891885, 0.763495418941602, 0.712142110569403, 0.0385209275409579, 0.96313498984091, 0.593413677765056, 0.744207103736699, 0.871114922920242, 0.0198914003558457, 0.784059509634972, 0.202999097527936, 0.390856014564633, 0.300481381127611, 0.229789084754884, 0.905745088588446, 0.739239115500823, 0.946140465559438, 0.314154316671193, 0.816592521732673, 0.877794414293021, 0.983992801746354, 0.265623347600922, 0.930095587391406, 0.745417301775888, 0.425958760781214, 0.929154467536137, 0.271716351853684, 0.643902056850493, 0.43678669584915, 0.619581998558715, 0.827017467701808, 0.531830659601837, 0.422172043006867, 0.81291391607374, 0.885273038875312, 0.309956403216347, 0.186331883771345, 0.424461955670267, 0.997192027978599, 0.433215622091666, 0.452280472731218, 0.69450547080487, 0.628442403394729, 0.884239240549505, 0.796065025497228, 0.649574688868597, 0.342127669602633, 0.937653557630256, 0.831226553069428, 0.668054030742496, 0.586903089890257, 0.653464525006711, 0.894822221482173, 0.53280886169523, 0.21522277360782, 0.957562289666384, 0.366296691354364, 0.491172014735639, 0.253177537582815, 0.57776112575084, 0.793586152372882, 0.540609549963847, 0.651492978911847, 0.812400965951383, 0.953959342325106, 0.788006828166544, 0.548453774768859, 0.0699334151577204, 0.83093127515167, 0.803405686514452, 0.395154554164037, 0.736415318679065, 0.473849432310089, 0.906922338763252, 0.0276497716549784, 0.925568751990795, 0.432583519956097, 0.118090635631233, 0.479757260065526, 0.73711645253934, 0.303509585792199, 0.977586327586323, 0.492351786233485, 0.927058070898056, 0.761865817243233, 0.291655492503196, 0.882443290902302, 0.0307668326422572, 0.285073815612122, 0.925895061343908, 0.782691380707547, 0.76410397561267, 0.655569962924346, 0.520098101347685, 0.016817229334265, 0.553733763750643, 0.702765831723809, 0.166872928384691, 0.69013331620954, 0.72312833275646, 0.293025034945458, 0.286745279794559, 0.201606729300693, 0.121016146382317, 0.516038313275203, 0.74231781065464, 0.178189663216472, 0.507309161126614, 0.485199409304187, 0.985329525079578, 0.900720057776198, 0.47163472417742, 0.982472016010433, 0.706687352852896, 0.0406438063364476, 0.141225004801527, 0.0619880568701774, 0.228569244267419, 0.855202492093667, 0.576731822453439, 0.506079103564844, 0.516747151734307, 0.321804024279118, 0.366600210778415, 0.928786049364135, 0.795964983059093, 0.244217006023973, 0.500293523306027, 0.146698108175769, 0.2838991493918, 0.718477793736383, 0.946046858560294, 0.0112665025517344, 0.814478608546779, 0.641478237695992, 0.408387447241694, 0.760034042410553, 0.303923466941342, 0.975880003999919, 0.616203985642642, 0.424227031646296, 0.0278202299959958, 0.36144590517506, 0.420723990071565, 0.87377503188327, 0.0100196597632021, 0.699814529856667, 0.674795751227066, 0.377566359937191, 0.917667685309425, 0.745123863685876, 0.322733566397801, 0.00818594638258219, 0.497275890782475, 0.00964190089143813, 0.195606124820188, 0.113144153729081, 0.63887476734817, 0.751869794679806, 0.815513349836692, 0.479083291254938, 0.155400038463995, 0.878198734950274, 0.133838660782203, 0.349681800464168, 0.421380625804886, 0.172857170924544, 0.476307974895462, 0.863634834298864, 0.762908318312839, 0.760517594637349, 0.70733231655322, 0.173426271183416, 0.858771594939753, 0.752761952811852, 0.400221461430192, 0.499806065810844, 0.576762856217101, 0.343484715092927, 0.0975290453061461, 0.754847318166867, 0.669034844264388, 0.212953122099862, 0.485223518218845, 0.667572805657983, 0.930634724907577, 0.74778698454611, 0.498949313536286, 0.644018102204427, 0.747846882557496, 0.83958694152534, 0.00684828707017004, 0.800260178511962, 0.433177990373224, 0.946352966595441, 0.741824764292687, 0.0714002479799092, 0.395703181624413, 0.562553631840274, 0.508152064168826, 0.368376362835988, 0.126659939996898, 0.948286331724375, 0.69894331600517, 0.0826302296482027, 0.117573478724808, 0.411453958135098, 0.830875678220764, 0.680840001674369, 0.966474500717595, 0.506051750155166, 0.830546265235171, 0.688026971183717, 0.423456026008353, 0.637592174578458, 0.869571596616879, 0.340672064339742, 0.617039978504181, 0.0489312615245581, 0.145921784453094, 0.344593204790726, 0.971291170688346, 0.460326031316072, 0.991608701646328, 0.710247130366042, 0.0351902977563441, 0.0926297714468092, 0.866059040883556, 0.608021950582042, 0.945616579148918, 0.255125740310177, 0.286997072631493, 0.33694353653118, 0.941623548744246, 0.0398013545200229, 0.0921582092996687, 0.89649895648472, 0.0835492901969701, 0.588910808553919, 0.647165805334225, 0.504609845578671, 0.691384450299665, 0.522524417145178, 0.840462482068688, 0.527872825274244, 0.752246274147183, 0.867788608185947, 0.331511733820662, 0.466354372445494, 0.329309717286378, 0.0989501194562763, 0.347737275762483, 0.869494940619916, 0.86616858211346, 0.545163782080635, 0.974596058949828, 0.415511773899198, 0.298362748231739, 0.578544655814767, 0.821085390169173, 0.797145596938208, 0.290876161539927, 0.73384153121151, 0.997169421985745, 0.489467493491247, 0.342167333466932, 0.525718363700435, 0.227671595755965, 0.144303075037897, 0.795623745536432, 0.589459720999002, 0.177346367156133, 0.133622613735497, 0.651839948724955, 0.459565082797781, 0.719245525309816, 0.945310077629983, 0.765306677669287, 0.314858115045354, 0.133905176073313, 0.254755511181429, 0.712024630280212, 0.109732352662832, 0.789245701394975, 0.456978281261399, 0.203608926385641, 0.859725995222107, 0.610425938153639, 0.314228063682094, 0.202379435533658, 0.941447838209569, 0.541162739740685, 0.0572077534161508, 0.965013519395143, 0.4918943264056, 0.980374098755419, 0.254936783341691, 0.635729317553341, 0.351542662130669, 0.30880323657766, 0.260982461972162, 0.475820215651765, 0.602634291863069, 0.181690144585446, 0.676044136052951, 0.481361341895536, 0.193233594065532, 0.0666118608787656, 0.931735865073279, 0.0634628471452743, 0.247105525806546, 0.0489637141581625, 0.0432545430958271, 0.402170493733138, 0.99500083271414, 0.982950035948306, 0.919892932521179, 0.27984995697625, 0.599682593951002, 0.00387653755024076, 0.702515915269032, 0.429796005599201, 0.608964761951938, 0.0873257890343666, 0.455451736692339, 0.417954809265211, 0.489048621384427, 0.256551993545145, 0.987192805390805, 0.915121516678482, 0.645457087550312, 0.817481776000932, 0.815550423460081, 0.490053222514689, 0.0501299276947975, 0.322315315948799, 0.250712374225259, 0.749668921111152, 0.572608456015587, 0.57366179372184, 0.949230512604117, 0.0592081889044493, 0.259007861604914, 0.26476241950877, 0.00258530792780221, 0.259104379452765, 0.176584523869678, 0.722977744415402, 0.398317599436268, 0.731817527441308, 0.531640341738239, 0.678343659266829, 0.123915693489835, 0.196715406607836, 0.724496683338657, 0.194525714498013, 0.817262659780681, 0.0651730631943792, 0.908251078100875, 0.144221723778173, 0.33783045806922, 0.509313893038779, 0.910683843074366, 0.776057346491143, 0.365020936820656, 0.820825756061822, 0.419154979754239, 0.329012283589691, 0.0491775486152619, 0.2268951383885, 0.4766952523496, 0.229053724557161, 0.764667161274701, 0.895998995518312, 0.806615548906848, 0.500313895056024, 0.50642124726437, 0.0290504368022084, 0.460201820824295, 0.654786457074806, 0.282219096785411, 0.0735413262154907, 0.125640891492367, 0.562467444920912, 0.958512428682297, 0.309521587332711, 0.442875719629228, 0.310983966104686, 0.018552357563749, 0.841467217775062, 0.485638073412701, 0.71223357366398, 0.178670806111768, 0.971054522320628, 0.516055568587035, 0.344512446783483, 0.649980801623315, 0.0659583632368594, 0.299740715883672, 0.895002889912575, 0.76063013356179, 0.339072461705655, 0.447045781416818, 0.375827538315207, 0.337679531425238, 0.585968421772122, 0.442641584668308, 0.661282954039052, 0.900973321637139, 0.0581104510929435, 0.34528957773, 0.906126101966947, 0.283140373183414, 0.468157862313092, 0.919857148779556, 0.911663803271949, 0.171567454468459, 0.00107277580536902, 0.430658097844571, 0.432033088756725, 0.664600017713383, 0.967515619471669, 0.128438560757786, 0.0443723746575415, 0.648558543529361, 0.56715228012763, 0.780567300738767, 0.113682709634304, 0.666645962512121, 0.845751892309636, 0.404675233876333, 0.720995582873002, 0.894102644640952, 0.671022681053728, 0.544814808294177, 0.546494105830789, 0.119247001595795, 0.378268601605669, 0.657870170427486, 0.686621298315004, 0.193299516802654, 0.363257724093273, 0.83664378686808, 0.992102030664682, 0.254711143439636, 0.465608331607655, 0.76733788033016, 0.672150928294286, 0.613128659548238, 0.535270302323624, 0.690596589120105, 0.138743097661063, 0.36174331814982, 0.550333178369328, 0.566434203181416, 0.0803885129280388, 0.791012504370883, 0.314464347204193, 0.147602481069043, 0.324321569874883, 0.400101883569732, 0.319017075467855, 0.180799973197281, 0.515828695846722, 0.66755224391818, 0.912194936536252, 0.176994561683387, 0.204459631582722, 0.500934184528887, 0.234562857775018, 0.0947971346322447, 0.292107390007004, 0.721380941336975, 0.549996228888631, 0.847179643809795, 0.902111523784697, 0.963936847168952, 0.627935978583992, 0.532280908664688, 0.146177751943469, 0.675771434325725, 0.241333447396755, 0.490630312822759, 0.10920449718833, 0.900257595581934, 0.589849884388968, 0.998135359724984, 0.0650957257021219, 0.481551870936528, 0.121677880408242, 0.30103305215016, 0.729538504267111, 0.116259658476338, 0.499301260570064, 0.919936014106497, 0.574295367579907, 0.612324612913653, 0.107494013151154, 0.628022098448128, 0.623016945552081, 0.836373091908172, 0.607209138572216, 0.963766561122611, 0.684681311249733), actual = c(1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"