Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.7210338        0.7210338 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5503662        0.5503662        0.5047263 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.48588365665637, 0.199859228916466, 0.139034040039405, 0.997812498360872, 0.0423778304830194, 0.0833675656467676, 0.0812862841412425, 0.785769386216998, 0.865755954524502, 0.821859920397401, 0.549503329908475, 0.0553037151694298, 0.91307840612717, 0.96980603504926, 0.857629874954, 0.796480790479109, 0.0812241258099675, 0.804788279812783, 0.732318199006841, 0.970322128385305, 0.402036346495152, 0.124081350630149, 0.525053040590137, 0.830363647080958, 0.505970390280709, 0.00885395612567663, 0.687104368582368, 0.194263640325516, 0.683389292098582, 0.390386193757877, 0.866102874279022, 0.439385053468868, 0.877949101384729, 0.922463586553931, 0.498649569926783, 0.872612603241578, 0.330836549168453, 0.235802470240742, 0.597119651036337, 0.28285378892906, 0.915511215105653, 0.651079135015607, 0.107910836813971, 0.421406658133492, 0.254810118349269, 0.946825133170933, 0.338525236817077, 0.514231787296012, 0.269672023830935, 0.548243465367705, 0.207887995289639, 0.987750204978511, 0.396563863381743, 0.745934020960703, 0.10227607190609, 0.456400375114754, 0.710658942814916, 0.935342885088176, 0.584505639271811, 0.992429001256824, 0.0584763549268246, 0.719335747184232, 0.391823170240968, 0.0301104926038533, 0.15812942921184, 0.742520675761625, 0.985268988879398, 0.605505974730477, 0.110460949363187, 0.916350695770234, 0.951925518922508, 0.653908435953781, 0.145492331357673, 0.229006846435368, 0.45990348700434, 0.634413515217602, 0.721193416975439, 0.378033920191228, 0.272100744536147, 0.219601613469422, 0.330812589498237, 0.59978354582563, 0.436281081754714, 0.00959470355883241, 0.829341441625729, 0.536126746563241, 0.0843040675390512, 0.915082082850859, 0.502253338694572, 0.845136649440974, 0.555733716581017, 0.948847248684615, 0.35402772622183, 0.909075196366757, 0.861774585675448, 0.180135764880106, 0.558669778518379, 0.601473461138085, 0.997258188668638, 0.982879745773971, 0.924398407805711, 0.695638871751726, 0.845672027207911, 0.320916190743446, 0.219144805334508, 0.0852371798828244, 0.645638308487833, 0.399009513203055, 0.536137391580269, 0.89271012833342, 0.0869759530760348, 0.787820717319846, 0.243875615065917, 0.270772258052602, 0.293478690786287, 0.83694420917891, 0.0233916237484664, 0.71904187859036, 0.568131518550217, 0.36005789344199, 0.731870768591762, 0.19646611995995, 0.865630874643102, 0.200415341416374, 0.367123297881335, 0.52396475081332, 0.259600583463907, 0.163110605906695, 0.357636958127841, 0.881895282072946, 0.707686237292364, 0.0705570545978844, 0.445749490289018, 0.0717904488556087, 0.294142711674795, 0.0853853318840265, 0.544179955963045, 0.418573544593528, 0.114467734936625, 0.640451579820365, 0.580497786868364, 0.305119099328294, 0.485687995329499, 0.813056929036975, 0.690890948055312, 0.782480418682098, 0.664724086178467, 0.25231333845295, 0.0163180767558515, 0.887506340630352, 0.608223133487627, 0.739017013693228, 0.396004005800933, 0.0649361885152757, 0.157002931926399, 0.0463814986869693, 0.257295456947759, 0.973355914233252, 0.719292283058167, 0.65408042864874, 0.660874441731721, 0.685882347403094, 0.442543738521636, 0.396866257768124, 0.758347108028829, 0.482370611978695, 0.910876613110304, 0.367654871894047, 0.900001079309732, 0.417165660997853, 0.36629467154853, 0.316497553838417, 0.217154492856935, 0.676642199512571, 0.863879770040512, 0.519510062178597, 0.0394287253729999, 0.77770545007661, 0.269656711490825, 0.880123218055815, 0.115851839538664, 0.640093418303877, 0.526550050126389, 0.996743158437312, 0.123833263525739, 0.194396109785885, 0.558441376313567, 0.0885971239767969, 0.203870846191421, 0.0327293039299548, 0.363517021061853, 0.670063707977533, 0.951403669314459, 0.810651367995888, 0.0473329771775752, 0.172737402375787, 0.191665389109403, 0.740575061878189, 0.479722995078191, 0.331483808113262, 0.724075510166585, 0.681001727934927, 0.958058215910569, 0.557666530366987, 0.266908994410187, 0.759512178367004, 0.121661227662116, 0.898347198031843, 0.349454975919798, 0.6890002861619, 0.456015764735639, 0.229176073800772, 0.280408863676712, 0.924994763918221, 0.678290990414098, 0.801875963807106, 0.327400868292898, 0.515818763291463, 0.0290176144335419, 0.622975016012788, 0.00804770831018686, 0.714965301798657, 0.495679121930152, 0.922471361234784, 0.569410369032994, 0.601131811272353, 0.1183255254291, 0.199133390327916, 0.909259726526216, 0.567494510672987, 0.304698549211025, 0.509988074889407, 0.528676232788712, 0.0787988875526935, 0.620224266545847, 0.203424282604828, 0.385322978254408, 0.830709960544482, 0.404659335501492, 0.59393453435041, 0.12050986639224, 0.34439314622432, 0.0931529379449785, 0.708398473449051, 0.200246569234878, 0.734966641059145, 0.15458827978, 0.944008068181574, 0.671287114964798, 0.0961608861107379, 0.192022369708866, 0.45536946086213, 0.604694434907287, 0.528874318581074, 0.00515503156930208, 0.748722786549479, 0.318350854795426, 0.379016017541289, 0.791644894052297, 0.698217313271016, 0.208881963975728, 0.794359144056216, 0.980286760721356, 0.223107347497717, 0.0256604885216802, 0.479375913972035, 0.746099677402526, 0.0399452038109303, 0.102982258889824, 0.434144004248083, 0.117732803802937, 0.743784582940862, 0.155679198447615, 0.962436678120866, 0.838220879901201, 0.671928432770073, 0.970068981405348, 0.382086648838595, 0.695817393483594, 0.343685885891318, 0.184559950139374, 0.99935534899123, 0.0671996036544442, 0.914112772792578, 0.20184420584701, 0.691253845114261, 0.810213614488021, 0.630743103800341, 0.292790514882654, 0.793724482646212, 0.260171551955864, 0.023574658902362, 0.287185663590208, 0.516698418185115, 0.480036458233371, 0.776063728379086, 0.0214393383357674, 0.51347229979001, 0.65131268161349, 0.0227227965369821, 0.918495843652636, 0.165523508097976, 0.0815178367774934, 0.0545466695912182, 0.943683887599036, 0.82790181459859, 0.654426580527797, 0.811728472588584, 0.208969477098435, 0.858193950494751, 0.437877187272534, 0.699706075014547, 0.787035320885479, 0.209650939563289, 0.747416350990534, 0.73075874010101, 0.886237833183259, 0.0711340380366892, 0.0609619950409979, 0.341428036568686, 0.0611108660232276, 0.562987753888592, 0.260928064817563, 0.512244732351974, 0.999871876556426, 0.00115943956188858, 0.809407770633698, 0.533432676456869, 0.194191637681797, 0.0627414090558887, 0.29676113370806, 0.435970978811383, 0.86253963294439, 0.0841911702882499, 0.952931062085554, 0.235889416653663, 0.154053839156404, 0.251891906838864, 0.169552801642567, 0.0353808917570859, 0.674030640628189, 0.565580911235884, 0.540043640648946, 0.474299005232751, 0.120728961657733, 0.443872033618391, 0.147291823057458, 0.359398865839466, 0.820097344228998, 0.570330874063075, 0.894817653112113, 0.435114352032542, 0.795475775841624, 0.646683198865503, 0.974414230789989, 0.227638400858268, 0.659426429774612, 0.540759506402537, 0.718855125829577, 0.666088874451816, 0.0370346137788147, 0.31051775556989, 0.72095869272016, 0.0139703927561641, 0.784300955478102, 0.785309257451445, 0.994585932930931, 0.963628117693588, 0.754824845818803, 0.822309371782467, 0.65882018674165, 0.275540123693645, 0.0151918444316834, 0.681115586310625, 0.34178065112792, 0.585731977364048, 0.118204342201352, 0.718746955739334, 0.790044502587989, 0.541747029870749, 0.858027010690421, 0.418115940177813, 0.993346168193966, 0.164552245987579, 0.381124335341156, 0.0805413245689124, 0.290494868997484, 0.266779084457085, 0.596464254194871, 0.33989100670442, 0.766119591658935, 0.520129388431087, 0.690185827435926, 0.901148781878874, 0.438709345646203, 0.0479410947300494, 0.214556081919, 0.714428456733003, 0.781974386423826, 0.165849711978808, 0.712342717917636, 0.695936291245744, 0.178984398255125, 0.960348235908896, 0.897972851293162, 0.446148515678942, 0.0136479744687676, 0.950890904059634, 0.0720919989980757, 0.789159052073956, 0.145522093400359, 0.682129892520607, 0.510888326447457, 0.405807442730293, 0.521797339199111, 0.271732467459515, 0.363877291791141, 0.891151458723471, 0.364510191371664, 0.11697278986685, 0.797229643911123, 0.457355552585796, 0.405219168635085, 0.774313564645126, 0.625300547340885, 0.917177906492725, 0.865771163254976, 0.999435727717355, 0.0463879154995084, 0.0219358841422945, 0.824695631163195, 0.342989081516862, 0.929119619075209, 0.180596065474674, 0.368456901516765, 0.134269945556298, 0.623230956494808, 0.672117538983002, 0.417635827558115, 0.00222058081999421, 0.889363375492394, 0.562290575355291, 0.416212514508516, 0.34346143482253, 0.0715960897505283, 0.0538714395370334, 0.227667194558308, 0.525395337259397, 0.741350332507864, 0.304500161670148, 0.903257236350328, 0.969392284285277, 0.854158839210868, 0.0244412908796221, 0.817702650791034, 0.526729087345302, 0.458206124603748, 0.801152071217075, 0.669722964754328, 0.73953760555014, 0.013135012704879, 0.581469275057316, 0.217938742367551, 0.51536246878095, 0.714454033412039, 0.0565206988248974, 0.684700807789341, 0.573358295252547, 0.427146887406707, 0.769685540115461, 0.614871296333149, 0.460876687662676, 0.266649647150189, 0.564121443312615, 0.86888579512015, 0.927709782961756, 0.0328150552231818, 0.91828732425347, 0.955804259749129, 0.229432683903724, 0.332957028178498, 0.322090920060873, 0.979988003848121, 0.489589106524363, 0.897148850141093, 0.109701755456626, 0.035777093609795, 0.348007073625922, 0.0119681512005627, 0.532184784300625, 0.510396310593933, 0.909415573813021, 0.957776399329305, 0.464573659002781, 0.217457194114104, 0.965614762390032, 0.886148139368743, 0.613389192614704, 0.486613756977022, 0.172526479000226, 0.0841931309551001, 0.200272541958839, 0.110805061645806, 0.197679692180827, 0.441586958710104, 0.273572134552523, 0.168547327863052, 0.737326528411359, 0.835666437866166, 0.790517077548429, 0.445678534917533, 0.212257049046457, 0.930871425196528, 0.071507012937218, 0.41819893126376, 0.808863553451374, 0.424810177180916, 0.989680773578584, 0.379690520465374, 0.232386440038681, 0.177541888318956, 0.568416576366872, 0.0493642007932067, 0.220797811867669, 0.681224277010188, 0.826296398183331, 0.302552819717675, 0.611593509092927, 0.23584016575478, 0.142192418919876, 0.00809330819174647, 0.0983271934092045, 0.586188288871199, 0.200858971336856, 0.272867136867717, 0.600341255078092, 0.510267154779285, 0.486305142985657, 0.238685836317018, 0.169164198683575, 0.117258935468271, 0.546425333013758, 0.171498344745487, 0.876865330850706, 0.305919965496287, 0.699273833306506, 0.350321494042873, 0.637788381893188, 0.714554666308686, 0.0578160958830267, 0.325855759903789, 0.66577562969178, 0.454421688104048, 0.609220580663532, 0.999356726417318, 0.770616515772417, 0.303530029486865, 0.332286562770605, 0.203811464132741, 0.0699331124778837, 0.261245897738263, 0.124045626958832, 0.801256423117593, 0.243599848588929, 0.191436200635508, 0.699470208259299, 0.841498749097809, 0.809658238198608, 0.282787554431707, 0.424398820847273, 0.293685347773135, 0.810370811494067, 0.647305053658783, 0.106218605069444, 0.326966151595116, 0.856735486304387, 0.956263640662655, 0.810179564636201, 0.519305235473439, 0.379187833284959, 0.460039121564478, 0.146347888745368, 0.275134440744296, 0.480840807082132, 0.518902747193351, 0.828180911485106, 0.727546923561022, 0.470506933284923, 0.742381460499018, 0.1167071480304, 0.564736023312435, 0.629391908412799, 0.888522386085242, 0.326968115055934, 0.196139967767522, 0.141154924640432, 0.232345784315839, 0.58523021871224, 0.095335055841133, 0.356513096718118, 0.146707599516958, 0.731017956510186, 0.211488489527255, 0.46074046427384, 0.228134316857904, 0.970226364443079, 0.400381781626493, 0.185929803177714, 0.874756697565317, 0.313174089882523, 0.430932842427865, 0.398458653828129, 0.980642119888216, 0.5287998507265, 0.121449662838131, 0.0549622965045273, 0.100040026009083, 0.755628663580865, 0.0616864070761949, 0.144500171998516, 0.17768522631377, 0.957163288490847, 0.719959396170452, 0.30544270738028, 0.721230092924088, 0.694081597961485, 0.780980352545157, 0.645822390448302, 0.78347758250311, 0.157614891417325, 0.0320346525404602, 0.0596350864507258, 0.14529690076597, 0.735248105367646, 0.920136277331039, 0.8260651929304, 0.705695025622845, 0.378515590913594, 0.461835100781173, 0.827159674372524, 0.894820854999125, 0.498730960767716, 0.859521902864799, 0.692158024059609, 0.124018958304077, 0.138320943107828, 0.848034220049158, 0.693885633489117, 0.461887924233451, 0.0971864969469607, 0.101218514144421, 0.352922245394439, 0.835259517189115, 0.326594151556492, 0.589604951441288, 0.994081704178825, 0.232899513095617, 0.746535482816398, 0.85536957369186, 0.34133190754801, 0.38482414977625, 0.808027347549796, 0.0456501001026481, 0.0838019179645926, 0.922751820879057, 0.909417971037328, 0.727115539833903, 0.812177937477827, 0.539254911243916, 0.0647154641337693, 0.195771113969386, 0.814047742867842, 0.374165730783716, 0.520165183348581, 0.380100751994178, 0.877844482660294, 0.728763096500188, 0.665329557843506, 0.942393891746178, 0.831091000931337, 0.429970704950392, 0.367477265186608, 0.383641498396173, 0.463935728184879, 0.85174502665177, 0.327615726273507, 0.67656484968029, 0.274761467473581, 0.153278273297474, 0.295074647059664, 0.430438974872231, 0.149640000425279, 0.0481059639714658, 0.108187919249758, 0.772047983948141, 0.278294504620135, 0.431233684066683, 0.497829029103741, 0.363872484536842, 0.434269825695083, 0.459260774310678, 0.841891370248049, 0.0857230324763805, 0.491107616340742, 0.0150254694744945, 0.266521226381883, 0.905537922400981, 0.630293004680425, 0.675104568246752, 0.807585075730458, 0.566487849922851, 0.494038912700489, 0.793333143228665, 0.939372485037893, 0.526096713962033, 0.448076093336567, 0.130307625280693, 0.650334337260574, 0.473716273205355, 0.662378297653049, 0.864540215348825, 0.473437688080594, 0.129635193152353, 0.0739753248635679, 0.94927660911344, 0.362310701282695, 0.866109624737874, 0.587324125925079, 0.666161267086864, 0.668076719157398, 0.806000411044806, 0.675076916115358, 0.0209949528798461, 0.786705454811454, 0.839478387730196, 0.595495764166117, 0.169269648846239, 0.00541308824904263, 0.174049036577344, 0.513805098598823, 0.431443984154612, 0.156861482420936, 0.705060650827363, 0.0844562263227999, 0.800867963349447, 0.959424297325313, 0.8679628088139, 0.79142814851366, 0.824856305029243, 0.526166788768023, 0.0517518324777484, 0.570626945467666, 0.119989589089528, 0.200136220781133, 0.60102042555809, 0.689224518137053, 0.524012468289584, 0.549003542866558, 0.025948827387765, 0.848569943569601, 0.467962696915492, 0.406122936634347, 0.281609000172466, 0.777033921098337, 0.351859759772196, 0.225295425858349, 0.617031645495445, 0.519022376509383, 0.880741458851844, 0.749534383881837, 0.0705148263368756, 0.452677921392024, 0.14548105141148, 0.844083341071382, 0.180977914715186, 0.34219119604677, 0.274367896141484, 0.574933343101293, 0.634179475484416, 0.40029480191879, 0.49345602071844, 0.915710494620726, 0.955210861051455, 0.705839595524594, 0.0939483141992241, 0.92317084595561, 0.840118204709142, 0.844846108928323, 0.614567008800805, 0.488052731845528, 0.696671446086839, 0.394510704791173, 0.142078248085454, 0.224415173754096, 0.302878431044519, 0.289970304118469, 0.80916082020849, 0.509895712137222, 0.351725872373208, 0.801058443263173, 0.0449165932368487, 0.463936613406986, 0.475591920316219, 0.136912422487512, 0.627381059806794, 0.718447773251683, 0.491239522816613, 0.816408502869308, 0.440525858197361, 0.853594517568126, 0.302361718611792, 0.530907924287021, 0.556455870857462, 0.67086537135765, 0.62457190034911, 0.494565542088822, 0.521028510294855, 0.992040687706321, 0.254238136345521, 0.379406232619658, 0.00618248619139194, 0.939231707947329, 0.579338226467371, 0.156497750198469, 0.758189542451873, 0.308189630974084, 0.507428295910358, 0.997349340235814, 0.80106820515357, 0.880815263837576, 0.821832457091659, 0.647791786585003, 0.251589379739016, 0.026371355401352, 0.519507498014718, 0.689756404841319, 0.110249071381986, 0.533550820779055, 0.533373759128153, 0.00449371919967234, 0.780633023940027, 0.349030191311613, 0.753284450853243, 0.433583088684827, 0.0993668455630541, 0.809769481420517, 0.322087070671842, 0.689860242651775, 0.264156248886138, 0.708381711272523, 0.802212370792404, 0.781145910266787, 0.980993306264281, 0.21856721653603, 0.0277266050688922, 0.669380761450157, 0.192266386700794, 0.655909022549167, 0.935229460243136, 0.871572556206957, 0.422543894033879, 0.423035238403827, 0.334298291010782, 0.077191689517349, 0.254859461681917, 0.862866642186418, 0.756163502577692, 0.902172827394679, 0.628403102047741, 0.933660526992753, 0.599984040716663, 0.551256045466289, 0.286645098589361, 0.561275937128812, 0.31668049050495, 0.22257191198878, 0.256038647610694, 0.0387896317988634, 0.89100576355122, 0.597294148057699, 0.180200944421813, 0.537455178098753, 0.250927023589611, 0.712096020579338, 0.242030690424144, 0.77062067296356, 0.976984948851168, 0.186297981766984, 0.669590669218451, 0.0675309875514358, 0.492790719959885, 0.658620192203671, 0.683096769033, 0.649558557430282, 0.901959192706272, 0.755109596531838, 0.570776669541374, 0.55752976774238, 0.11468508373946, 0.230720392894, 0.448268725769594, 0.211943504866213, 0.758105408400297, 0.219272059155628, 0.713129833107814, 0.867013603681698, 0.491948768496513, 0.0693643125705421, 0.262410232098773, 0.546002758899704, 0.241714218165725, 0.121849432820454, 0.277497579809278, 0.677472590235993, 0.455542211653665, 0.917802513809875, 0.719167427858338, 0.28661164874211, 0.374088914133608, 0.486838996643201, 0.552533959737048, 0.265318232355639, 0.0970573958475143, 0.584772074362263, 0.827791502932087, 0.987592789810151, 0.661044236505404, 0.408515138784423, 0.610074014170095, 0.0425440901890397, 0.114281221525744, 0.35715526342392, 0.163061921019107, 0.677534633083269, 0.793599510332569, 0.631322807632387, 0.656429550610483, 0.0745702914427966, 0.324336330639198, 0.143270685570315, 0.635680892970413, 0.595405055209994, 0.384355754358694, 0.985751824686304, 0.116390279261395, 0.251547212712467, 0.463069283869117, 0.441405104938895, 0.559189605293795, 0.73854945669882, 0.231948325876147, 0.751722207991406, 0.221579217584804, 0.258352431003004, 0.596706897253171, 0.329318854259327, 0.122805308550596, 0.338948193006217, 0.496464914176613, 0.8337319761049, 0.016749091912061, 0.127742408309132, 0.762237968621776, 0.538071306189522, 0.593669844092801, 0.0838415361940861, 0.218329237541184, 0.756133282789961, 0.986237879376858, 0.732391847064719, 0.919811060884967, 0.390018530655652, 0.0946428733877838, 0.252293780911714, 0.555286096641794, 0.674587741727009, 0.479686119128019, 0.684694505296648, 0.620435666758567, 0.174374315189198, 0.584959503961727, 0.104837810387835, 0.810631531989202, 0.157017733668908, 0.186591923236847, 0.182137674884871, 0.202971373219043, 0.226261936826631, 0.136641970835626, 0.105329910060391, 0.0599889357108623, 0.865357774542645, 0.13677887711674, 0.0383136800955981, 0.0149083139840513, 0.812080237315968, 0.158273113891482, 0.449403953040019, 0.28235229337588, 0.262400978943333), actual = c(1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.501138044521213, 0.862477088114247, 0.467392419232056, 0.542666810099036, 0.712994099128991, 0.890194173669443, 0.296075433492661, 0.597653770120814, 0.99740189826116, 0.875449479091913, 0.00475002848543227, 0.00911456206813455, 0.715021649608389, 0.700246908934787, 0.464053225005046, 0.0130996233783662, 0.623033637180924, 0.272652106592432, 0.285930621437728, 0.291029270272702, 0.579385793302208, 0.697694837581366, 0.0441094492562115, 0.939765635877848, 0.187246093992144, 0.109318409580737, 0.0324626211076975, 0.459495073882863, 0.358766355086118, 0.145008663181216, 0.649006966967136, 0.167698757024482, 0.50806796271354, 0.675202110083774, 0.24005648586899, 0.347446524305269, 0.419280972564593, 0.772769138915464, 0.877665085019544, 0.84736479097046, 0.0169450987596065, 0.0320826612878591, 0.430163377663121, 0.689437380526215, 0.490430038888007, 0.00161180947907269, 0.322118627373129, 0.662410457851365, 0.986559883225709, 0.271510052494705, 0.88122813240625, 0.647309295833111, 0.0903654047288001, 0.413718735566363, 0.537303790682927, 0.477715268731117, 0.318990790052339, 0.962659578770399, 0.693291049217805, 0.820301024941728, 0.240659341681749, 0.866147852968425, 0.315545206423849, 0.74361528409645, 0.137154101626948, 0.296562945703045, 0.163197732530534, 0.728038578061387, 0.849695373792201, 0.564674354624003, 0.703566267620772, 0.405717122135684, 0.194890706799924, 0.669494042405859, 0.0484612700529397, 0.67175219161436, 0.0409154361113906, 0.597004088340327, 0.576325064990669, 0.566403812961653, 0.346765984315425, 0.814104030840099, 0.216529776575044, 0.985362018691376, 0.011191857047379, 0.841021611122414, 0.849726446904242, 0.869671013206244, 0.59683420159854, 0.51267729466781, 0.174818044994026, 0.27069904981181, 0.708018193254247, 0.182798506924883, 0.970502105541527, 0.0440338959451765, 0.784616432385519, 0.61758716404438, 0.0313564699608833, 0.11677411547862, 0.532583184540272, 0.765328563284129, 0.4432194060646, 0.114764080382884, 0.678729796549305, 0.930670503526926, 0.133480905322358, 0.288971031783149, 0.421525764511898, 0.495446679182351, 0.0445210677571595, 0.35021915496327, 0.527577227680013, 0.391394688514993, 0.882195349084213, 0.325746770948172, 0.382827162044123, 0.644713802495971, 0.281627166783437, 0.998801517067477, 0.860386670334265, 0.645085159689188, 0.183346532750875, 0.881390021182597, 0.98536999640055, 0.903594758594409, 0.483834096696228, 0.580014869337901, 0.791767289163545, 0.481595461023971, 0.0297910664230585, 0.772194976685569, 0.793970226543024, 0.970325664849952, 0.199437863891944, 0.480532564921305, 0.934887628303841, 0.903178303036839, 0.216326477238908, 0.130350479856133, 0.95330603257753, 0.243553611449897, 0.831269921502098, 0.362924007233232, 0.383088808506727, 0.619190747616813, 0.159762183437124, 0.780668327584863, 0.53928501624614, 0.127684990875423, 0.653666608501226, 0.530215302016586, 0.826728232903406, 0.322887012036517, 0.12486072210595, 0.564790845382959, 0.942423755768687, 0.623021605191752, 0.569794841343537, 0.529803192475811, 0.636737566906959, 0.789063681382686, 0.0516712551470846, 0.751060949638486, 0.890953824622557, 0.571158959297463, 0.199940529884771, 0.267496703658253, 0.97267460054718, 0.703916247701272, 0.175779422745109, 0.0937599616590887, 0.232638006098568, 0.805574018741027, 0.515077874995768, 0.652465968159959, 0.305114840622991, 0.897263207472861, 0.198536833748221, 0.528975209221244, 0.143337497720495, 0.884579349542037, 0.454908184939995, 0.558993002399802, 0.624917385634035, 0.531341649591923, 0.339218054432422, 0.510070000775158, 0.897817613556981, 0.506216962588951, 0.626371629303321, 0.0457000429742038, 0.659409519517794, 0.228676960570738, 0.346850991947576, 0.408618971705437, 0.177691115764901, 0.067180764162913, 0.649292556103319, 0.0706808427348733, 0.125648130429909, 0.786034904187545, 0.779445336665958, 0.787581128068268, 0.27595977508463, 0.580848799087107, 0.758214890956879, 0.913164591649547, 0.356799619738013, 0.855603180127218, 0.353511621244252, 0.311252840794623, 0.255784543929622, 0.249874771106988, 0.308140566572547, 0.219246905529872, 0.99465558025986, 0.245388109702617, 0.000645935535430908, 0.185846262145787, 0.483255595900118, 0.286739367060363, 0.36328872689046, 0.550411685369909, 0.204011575086042, 0.670315515482798, 0.554525169776753, 0.412263008533046, 0.118124180473387, 0.788747947197407, 0.370101967127994, 0.198810366680846, 0.322812060592696, 0.297377816867083, 0.630591403692961, 0.199539446271956, 0.703793656313792, 0.204471996519715, 0.846714480780065, 0.00680800713598728, 0.414765901397914, 0.170837761135772, 0.314336673123762, 0.876661606598645, 0.446962411282584, 0.787562829442322, 0.641171629307792, 0.962850646814331, 0.848019480705261, 0.996251977514476, 0.140365365659818, 0.958845473593101, 0.318818244151771, 0.863994248677045, 0.634938765317202, 0.584338169544935, 0.844626465113834, 0.0215765659231693, 0.630496334983036, 0.666331014363095, 0.691800433909521, 0.845897585852072, 0.61790129262954, 0.862181355711073, 0.521201508585364, 0.505662361858413, 0.22928554425016, 0.690111676463857, 0.758500552736223, 0.28043351881206, 0.702111104503274, 0.84485908341594, 0.63227828219533, 0.921816726913676, 0.320763756055385, 0.868450697278604, 0.538496463093907, 0.699782865121961, 0.802214903058484, 0.941171594196931, 0.0143063804134727, 0.384420550893992, 0.888552946737036, 0.546975726261735, 0.561721502104774, 0.89785927743651, 0.259855615440756, 0.951296354178339, 0.360648122150451, 0.26993043301627, 0.780680306255817, 0.572173375636339, 0.688994543161243, 0.733170195482671, 0.643552525201812, 0.411165648372844, 0.410229446133599, 0.26932879560627, 0.946200549835339, 0.878561056451872, 0.74796740966849, 0.236000698525459, 0.360928405076265, 0.600601505255327, 0.213185907807201, 0.412658972432837, 0.7640826262068, 0.221712152473629, 0.616947324713692, 0.553195840213448, 0.352969722822309, 0.33924611331895, 0.0124722656328231, 0.103222633479163, 0.156260762130842, 0.216464415192604, 0.893165416317061, 0.847122805425897, 0.0778612387366593, 0.519062035949901, 0.303876444464549, 0.0154534738976508, 0.463476288830861, 0.873122701654211, 0.938042587833479, 0.948738452745602, 0.173613076563925, 0.645023390650749, 0.98461577296257, 0.888771747471765, 0.00170996319502592, 0.945603668224066, 0.994392951019108, 0.93024096195586, 0.10785919521004, 0.89218100416474, 0.510002960916609, 0.434992203954607, 0.364336762344465, 0.857874366687611, 0.470846497453749, 0.363090366357937, 0.934876007260755, 0.26122164260596, 0.954374660039321, 0.0344914617016912, 0.745741735445336, 0.470216503832489, 0.904523356584832, 0.94016498234123, 0.326536253560334, 0.950172630371526, 0.0970544547308236, 0.736143828369677, 0.567628065124154, 0.712787033291534, 0.84634636878036, 0.6322186216712, 0.421173910377547, 0.577899172669277, 0.142347481800243, 0.419279428664595, 0.893615298904479, 0.257541908416897, 0.52893482404761, 0.871405720710754, 0.368947862880304, 0.491995096672326, 0.928210954414681, 0.914565916871652, 0.386745525058359, 0.202546277781948, 0.599601605208591, 0.0406657487619668, 0.661344750085846, 0.0478967255912721, 0.540133715374395, 0.756139365024865, 0.418798421975225, 0.106756671564654, 0.416159985819831, 0.876512743066996, 0.559576787753031, 0.733054325450212, 0.361436887877062, 0.0222490965388715, 0.330803459975868, 0.855844227364287, 0.318482855800539, 0.450912116561085, 0.924965589540079, 0.607801403617486, 0.991449782159179, 0.521523267030716, 0.676896282006055, 0.434405853971839, 0.325081391492859, 0.664537999080494, 0.765572994248942, 0.677993192337453, 0.849771708715707, 0.16919637680985, 0.606392645975575, 0.711843604454771, 0.878507242305204, 0.640996899688616, 0.370605585165322, 0.585498807020485, 0.0736387609504163, 0.116129736648872, 0.324538840213791, 0.728400017367676, 0.00702859531156719, 0.823475958546624, 0.387813596986234, 0.408745745196939, 0.366736668394879, 0.0798864921089262, 0.964699888369069, 0.450575848342851, 0.556065980577841, 0.191776925232261, 0.748184072785079, 0.271667883265764, 0.603159321239218, 0.634245619177818, 0.27043362846598, 0.514144701184705, 0.118044181261212, 0.514401004882529, 0.919030906399712, 0.924826451810077, 0.961588556179777, 0.84454486425966, 0.47095829481259, 0.84780538524501, 0.929174948250875, 0.189639186020941, 0.592740436084569, 0.79013573541306, 0.198436981765553, 0.806281888857484, 0.395845412276685, 0.951281237881631, 0.907953994348645, 0.390699572861195, 0.501902868272737, 0.831976778805256, 0.161113040754572, 0.581216007471085, 0.466278214240447, 0.837837502360344, 0.963445558445528, 0.385240688920021, 0.181054728571326, 0.578338953666389, 0.0589719025883824, 0.561112317722291, 0.137384013738483, 0.638354057213292, 0.743224804056808, 0.462294925469905, 0.701407416490838, 0.959915824932978, 0.178857452236116, 0.865319098345935, 0.584905522176996, 0.064228558447212, 0.170073742745444, 0.600661578821018, 0.205536606023088, 0.789406371768564, 0.0945998553652316, 0.901438960572705, 0.57359438133426, 0.305806149030104, 0.192922450602055, 0.162587016588077, 0.118162415456027, 0.902705712476745, 0.846300726989284, 0.743636168772355, 0.7648933052551, 0.774611089145765, 0.491414377000183, 0.665865699294955, 0.417404243722558, 0.377837390638888, 0.360289533855394, 0.923928739735857, 0.391070971731097, 0.0463226346764714, 0.550201774341986, 0.604557842947543, 0.78120937733911, 0.569148062728345, 0.474968824069947, 0.237860373454168, 0.292980769183487, 0.184116386575624, 0.892034550895914, 0.515284545952454, 0.892432159744203, 0.271044144174084, 0.888308468274772, 0.966325473738834, 0.913592282682657, 0.809254665160552, 0.259171252604574, 0.763327957363799, 0.659184646094218, 0.489490239648148, 0.968011793447658, 0.0584150364156812, 0.0149538430850953, 0.364477637223899, 0.173210892826319, 0.447421161690727, 0.17281645629555, 0.403075337409973, 0.374498297227547, 0.225671576336026, 0.22082815063186, 0.959068156778812, 0.0315749940928072, 0.505732719786465, 0.532812477322295, 0.87850673077628, 0.321218176046386, 0.61631776811555, 0.213464847300202, 0.0696274721994996, 0.300640010507777, 0.433460290776566, 0.993128886446357, 0.23433374078013, 0.714359359117225, 0.442213755333796, 0.942326437681913, 0.607768931658939, 0.841278631240129, 0.209379391511902, 0.34638873883523, 0.399103078758344, 0.421608414733782, 0.52972628059797, 0.716441905824468, 0.288546537514776, 0.105318862013519, 0.246863641776145, 0.488592097535729, 0.198940492467955, 0.0700668692588806, 0.825931779108942, 0.629399961559102, 0.331343190046027, 0.318027023458853, 0.801810594741255, 0.00607322854921222, 0.659911973867565, 0.99288887809962, 0.636064171325415, 0.975126443430781, 0.94149679178372, 0.687678702874109, 0.660880613839254, 0.564283404964954, 0.472819137619808, 0.938676876481622, 0.56353893247433, 0.0617683618329465, 0.512593840714544, 0.488383320858702, 0.261148600606248, 0.555007735034451, 0.409416151698679, 0.49719176418148, 0.687544353771955, 0.769086970249191, 0.0673677858430892, 0.00855795107781887, 0.554312501335517, 0.899037217721343, 0.337359012104571, 0.898090806324035, 0.108920182101429, 0.151169738732278, 0.404120144899935, 0.974429630907252, 0.115786501206458, 0.687856900738552, 0.968193105654791, 0.924619155703112, 0.296071726828814, 0.7255207025446, 0.64095509191975, 0.547423495911062, 0.630570841487497, 0.776464147027582, 0.89387511764653, 0.664495565230027, 0.971570540685207, 0.138141200644895, 0.891205376945436, 0.834502110723406, 0.315131808631122, 0.72974732844159, 0.62834944576025, 0.274780781706795, 0.456817642785609, 0.688915011938661, 0.203334713354707, 0.163795131957158, 0.458711508661509, 0.62352415593341, 0.908188636414707, 0.379819781053811, 0.948145207250491, 0.199959008488804, 0.98861217382364, 0.723237130092457, 0.0394265323411673, 0.56863689632155, 0.571815472794697, 0.114900228334591, 0.0525951485615224, 0.372513412963599, 0.160547597100958, 0.0540953001473099, 0.754867784213275, 0.50841266149655, 0.15858341474086, 0.648767005652189, 0.208853326272219, 0.345834693172947, 0.102471374440938, 0.574272196739912, 0.846057612681761, 0.964369651395828, 0.115936845541, 0.604901482816786, 0.493239521048963, 0.869549304479733, 0.945594606455415, 0.266286288620904, 0.42749430402182, 0.75195736810565, 0.845800893148407, 0.310342744691297, 0.519510724814609, 0.193588313180953, 0.66619456675835, 0.720387725858018, 0.272583908867091, 0.796950380317867, 0.791165733477101, 0.655044556129724, 0.909204768482596, 0.162063726456836, 0.486466620815918, 0.107960410881788, 0.627154799876735, 0.28754210146144, 0.165729068918154, 0.490552082890645, 0.403234974713996, 0.917638863902539, 0.369189793476835, 0.603731156094, 0.0946886846795678, 0.997762226732448, 0.729404869722202, 0.229969158768654, 0.171226654201746, 0.353612883016467, 0.577916272450238, 0.810927937505767, 0.756866970798001, 0.881064702989534, 0.616810522973537, 0.171899719163775, 0.58353020157665, 0.806439850712195, 0.178891656454653, 0.152205326128751, 0.596456375671551, 0.825433819089085, 0.461854418739676, 0.448377031367272, 0.854972860077396, 0.602944620419294, 0.821401798631996, 0.292634807527065, 0.90237193997018, 0.455457285512239, 0.0798754000570625, 0.0934317554347217, 0.477749610319734, 0.0373683532234281, 0.613259318983182, 0.963915823260322, 0.287659255089238, 0.341204512165859, 0.511080038966611, 0.730174167081714, 0.327245380263776, 0.273481697309762, 0.0223942918237299, 0.600712012499571, 0.881484556710348, 0.711031807586551, 0.411871669813991, 0.132123944815248, 0.0732594865839928, 0.303574342280626, 0.498654979746789, 0.754717556759715, 0.494599475990981, 0.162625011289492, 0.366672305157408, 0.217400161083788, 0.186810737242922, 0.318622635211796, 0.162790064467117, 0.0319070583209395, 0.282744782743976, 0.317172698676586, 0.526879709912464, 0.76125977258198, 0.95825391355902, 0.324391333153471, 0.0030780581291765, 0.916599119780585, 0.647312471875921, 0.243295433232561, 0.767466541845351, 0.924795399187133, 0.94030737131834, 0.409748898120597, 0.464488343335688, 0.805421824101359, 0.815799101255834, 0.0697893786709756, 0.207752603571862, 0.283385089598596, 0.223660751711577, 0.607263281941414, 0.00182223925366998, 0.0114124324172735, 0.316251958487555, 0.500531634548679, 0.0188921974040568, 0.280291920760646, 0.957178918411955, 0.637533475412056, 0.260421651182696, 0.0171478737611324, 0.371466849464923, 0.910539963981137, 0.281235564965755, 0.215771335409954, 0.973489879397675, 0.258711460512131, 0.66244977992028, 0.994110639672726, 0.00726938666775823, 0.954534633085132, 0.0110235607717186, 0.555084869731218, 0.278867604676634, 0.119176703039557, 0.324777564033866, 0.630966571858153, 0.799363552359864, 0.111443962203339, 0.257263134000823, 0.0731619317084551, 0.125750122591853, 0.265138641931117, 0.682047665119171, 0.0820301538333297, 0.642910067923367, 0.492616239469498, 0.682355096796528, 0.304795351345092, 0.764329632511362, 0.571756955934688, 0.0074724534060806, 0.535672208294272, 0.969802533974871, 0.164862274425104, 0.00188520387746394, 0.843487042468041, 0.236203322187066, 0.647056013112888, 0.993883879156783, 0.613918354734778, 0.0992759654764086, 0.828319582622498, 0.404469420434907, 0.238960335729644, 0.175110630923882, 0.524167300667614, 0.110296875936911, 0.275789134670049, 0.0111969877034426, 0.883993057068437, 0.917761617572978, 0.196351701626554, 0.864021913846955, 0.27210669987835, 0.258215559879318, 0.333513117861003, 0.0397047677543014, 0.794484837446362, 0.542538099223748, 0.809866780648008, 0.474414737196639, 0.548660920234397, 0.959408639464527, 0.392225000774488, 0.777410451322794, 0.809844729956239, 0.498209147248417, 0.841101182624698, 0.479464319301769, 0.954503427725285, 0.292167330393568, 0.374607933685184, 0.16301609412767, 0.778947975952178, 0.952719998545945, 0.913585214177147, 0.478557122405618, 0.92172113340348, 0.184608693001792, 0.786301008425653, 0.229491173755378, 0.267642266815528, 0.0688436788041145, 0.046173692913726, 0.931063713040203, 0.849114158889279, 0.974042798858136, 0.576773702166975, 0.268410191638395, 0.8609696587082, 0.93831899156794, 0.623139902018011, 0.858783016446978, 0.877363916486502, 0.846362575655803, 0.199805428506806, 0.815473622642457, 0.21133491769433, 0.325724267866462, 0.859216082375497, 0.63973407773301, 0.36596155911684, 0.753632713807747, 0.421630820026621, 0.294078971957788, 0.453979470301419, 0.0770552204921842, 0.563401672756299, 0.922915368340909, 0.445827839430422, 0.952801384497434, 0.655369860352948, 0.577372292988002, 0.42427184805274, 0.242241232888773, 0.653809751616791, 0.75369264697656, 0.775038142688572, 0.231896995799616, 0.78069362603128, 0.203470538603142, 0.251842855010182, 0.332271856954321, 0.22540860879235, 0.61018795450218, 0.11329720611684, 0.762199755059555, 0.564786348259076, 0.877797955647111, 0.513966270023957, 0.742579312063754, 0.594946520635858, 0.99661481520161, 0.12830688408576, 0.0595844960771501, 0.773519914364442, 0.43301703222096, 0.208065640879795, 0.660296421730891, 0.437434859806672, 0.897605772828683, 0.554361160146073, 0.152781724696979, 0.192747961729765, 0.505507600028068, 0.0810436753090471, 0.541237014345825, 0.976283709984273, 0.463817247189581, 0.956813892349601, 0.807091361377388, 0.70449975435622, 0.147700921166688, 0.74853589409031, 0.28534435830079, 0.524071434512734, 0.632481631124392, 0.39944139868021, 0.400499936658889, 0.50395301124081, 0.489740632707253, 0.161274835234508, 0.144920125370845, 0.411381617188454, 0.773336787708104, 0.966753693064675, 0.394335728604347, 0.272963126888499, 0.221558183664456, 0.403492602752522, 0.71870670793578, 0.1406426995527, 0.435803540982306, 0.666500635445118, 0.734238299075514, 0.971161083783954, 0.79009275068529, 0.214339521713555, 0.866323451977223, 0.531363402726129, 0.01102499733679, 0.635329546639696, 0.773754663066939, 0.171397197060287, 0.105352928629145, 0.445194480707869, 0.864305064780638, 0.919440872268751, 0.971649902872741, 0.371432014042512, 0.850049703149125, 0.10276987310499, 0.579316496150568, 0.0745662732515484, 0.017229394055903, 0.617428267840296, 0.959256334463134, 0.00674310722388327, 0.317054592072964, 0.635846989694983, 0.812754975864664, 0.504282994894311, 0.552224125480279, 0.0315414050128311, 0.0586346082855016, 0.0151070463471115, 0.621936895651743, 0.775288480566815, 0.0658077716361731, 0.542458917247131, 0.433773668017238, 0.0332063236273825, 0.751639564521611, 0.630601003533229, 0.0796707575209439, 0.89560594339855, 0.872668822994456, 0.516456505050883, 0.15920915058814, 0.327874386683106, 0.561400577891618, 0.97468424658291, 0.0866923246067017, 0.979037587996572, 0.683732610894367, 0.873624287545681, 0.307236081454903, 0.458162672817707, 0.632776406593621, 0.431764479493722, 0.236402447102591, 0.482512650312856, 0.609719618223608, 0.694101883564144, 0.659644658211619, 0.438274104846641, 0.91644236119464, 0.243142229272053, 0.757142842980102, 0.729213340906426, 0.55458358861506, 0.53828401514329, 0.395753415999934), actual = c(0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"