Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.0734573        0.0734573 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5410912        0.5011224        0.5009640 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.230322209885344, 0.402366146212444, 0.58655693451874, 0.864713182672858, 0.800472818315029, 0.45040194154717, 0.689036699943244, 0.485063989181072, 0.185833313502371, 0.804569700267166, 0.979421446332708, 0.266100821318105, 0.645772344199941, 0.130996156716719, 0.286305011482909, 0.636715708533302, 0.689789519412443, 0.476405982160941, 0.278035054216161, 0.0184135071467608, 0.479195085354149, 0.0349199597258121, 0.337065741652623, 0.052432345924899, 0.117486144648865, 0.114529482554644, 0.240468897856772, 0.106038576457649, 0.878430165117607, 0.395716425264254, 0.510661759413779, 0.495987029979005, 0.757106459699571, 0.0694146046880633, 0.0147273470647633, 0.920672894222662, 0.370626560645178, 0.486472570104524, 0.269610984250903, 0.541865905281156, 0.136641931021586, 0.450559925520793, 0.965600557392463, 0.203141612699255, 0.178072372218594, 0.94672407140024, 0.787410984747112, 0.682579730870202, 0.260906137526035, 0.0604831783566624, 0.476762766251341, 0.024973911466077, 0.883081394946203, 0.950854825787246, 0.569586635334417, 0.468625674024224, 0.79617472877726, 0.651505801128224, 0.813935641897842, 0.265545705100521, 0.0718130045570433, 0.270046511199325, 0.626319223316386, 0.71460352698341, 0.69185327924788, 0.648690453497693, 0.244955029105768, 0.459949226351455, 0.452350896783173, 0.47922979388386, 0.39653759310022, 0.296468784566969, 0.748762601055205, 0.856355443131179, 0.266976830083877, 0.0784915594849735, 0.100096077425405, 0.0330305634997785, 0.183298569172621, 0.104563816916198, 0.779498826246709, 0.821008702507243, 0.815206829691306, 0.886974692577496, 0.55885222251527, 0.175622173352167, 0.00507078785449266, 0.341641181847081, 0.857831436675042, 0.868827425176278, 0.236388920806348, 0.95031627244316, 0.826383710838854, 0.249473104951903, 0.556936155306175, 0.0578270708210766, 0.567901095375419, 0.617541944840923, 0.732620309805498, 0.309025671565905, 0.126776096876711, 0.0325925210490823, 0.987103424500674, 0.382605773163959, 0.068670156178996, 0.16549524734728, 0.438081911532208, 0.793196263257414, 0.681854317896068, 0.529577085981146, 0.843530492624268, 0.881618598476052, 0.949329395778477, 0.930372360395268, 0.253017551731318, 0.527301792055368, 0.714278419269249, 0.689059871947393, 0.709351095603779, 0.763426349731162, 0.227825194131583, 0.325202939100564, 0.435970873571932, 0.975721318274736, 0.558247276116163, 0.950205254834145, 0.879930656868964, 0.993294064421207, 0.522080156952143, 0.181061412673444, 0.894181102514267, 0.635659364750609, 0.499663225375116, 0.638825125293806, 0.50588266691193, 0.783140338724479, 0.659292642725632, 0.195958487922326, 0.512695792131126, 0.386393165914342, 0.0240455875173211, 0.926640165969729, 0.14795504254289, 0.342317329254001, 0.425923677859828, 0.694605583557859, 0.484723669476807, 0.134494051570073, 0.296406128210947, 0.00260105915367603, 0.317088098265231, 0.34384989622049, 0.384934303816408, 0.923289683414623, 0.904067310038954, 0.126101040048525, 0.851010372629389, 0.346079312963411, 0.971545563545078, 0.204530849820003, 0.632221199106425, 0.779103388311341, 0.310617269948125, 0.849365649977699, 0.324621618492529, 0.224287641933188, 0.857348731486127, 0.322369665838778, 0.940917469793931, 0.520579257514328, 0.491202703444287, 0.158977757208049, 0.462201590184122, 0.602239415980875, 0.337011496303603, 0.632012191461399, 0.133649825584143, 0.747285968158394, 0.194625460542738, 0.345904546789825, 0.204811054980382, 0.905862483661622, 0.149446774739772, 0.819866482168436, 0.653724081814289, 0.761379294097424, 0.121182517614216, 0.501493075164035, 0.38390705967322, 0.938297016778961, 0.439850548282266, 0.710027780150995, 0.640845472924411, 0.971201267559081, 0.20915709529072, 0.915073270210996, 0.0414729376789182, 0.266804395476356, 0.496561839943752, 0.185681591276079, 0.124792031478137, 0.0745178817305714, 0.276776587590575, 0.298862520605326, 0.538579951040447, 0.941329214721918, 0.647213229443878, 0.491467230487615, 0.866098204161972, 0.612772318068892, 0.154565970879048, 0.65142653696239, 0.157656136667356, 0.118294185725972, 0.684978611534461, 0.0717976333107799, 0.608723749406636, 0.536730739986524, 0.90596369234845, 0.67553744232282, 0.282251528464258, 0.816335961921141, 0.904438181780279, 0.036319725215435, 0.754133127396926, 0.798172834794968, 0.270017408533022, 0.650580727495253, 0.50404055789113, 0.966945016756654, 0.214847256196663, 0.252348486101255, 0.667338790372014, 0.966074542375281, 0.72049209731631, 0.251153749180958, 0.704401971306652, 0.481330689275637, 0.0738749124575406, 0.0884901178069413, 0.360629029804841, 0.363876561867073, 0.273352260235697, 0.327266262145713, 0.80415975721553, 0.2411768801976, 0.862393742892891, 0.107587554259226, 0.785389120224863, 0.803366441978142, 0.248525835806504, 0.052900510141626, 0.6578334341757, 0.366641349857673, 0.220732851885259, 0.405446534277871, 0.226164725376293, 0.588499343954027, 0.0665041266474873, 0.023543082177639, 0.990798325277865, 0.811542451381683, 0.434749861247838, 0.618084751768038, 0.980753072770312, 0.394930668408051, 0.825995265506208, 0.244720629183576, 0.280398692702875, 0.168852224713191, 0.891047605779022, 0.907211171230301, 0.939866133732721, 0.212224826682359, 0.487407895736396, 0.65955959376879, 0.239133349386975, 0.379652498755604, 0.241131718503311, 0.418384582735598, 0.074032133910805, 0.129073417745531, 0.514492577174678, 0.5692594924476, 0.0495515109505504, 0.732840169686824, 0.244407567428425, 0.846795486286283, 0.0669304821640253, 0.429103180300444, 0.103913240134716, 0.77199849486351, 0.548015642212704, 0.892453544773161, 0.623788489028811, 0.114491700893268, 0.707207291154191, 0.282299456419423, 0.219484138768166, 0.484366024611518, 0.245052524842322, 0.691745514981449, 0.140236616600305, 0.49032486230135, 0.244657180272043, 0.752254183171317, 0.793767243856564, 0.12149410136044, 0.877265407238156, 0.351786813698709, 0.999118478270248, 0.133012520615011, 0.962312601273879, 0.374175944598392, 0.730315594933927, 0.00071215140633285, 0.92902046488598, 0.467043127631769, 0.793913387227803, 0.213806950254366, 0.834655485115945, 0.949067852925509, 0.428465848090127, 0.110212166327983, 0.703989563509822, 0.228656882187352, 0.323675619671121, 0.379779675742611, 0.483338589547202, 0.613338228547946, 0.51121197687462, 0.184693013783544, 0.333679223665968, 0.127224263036624, 0.693844134220853, 0.333544515771791, 0.81358456518501, 0.592697536572814, 0.645012214547023, 0.364354668650776, 0.184553828556091, 0.40800042799674, 0.691068469546735, 0.992840112186968, 0.481674146139994, 0.803197349188849, 0.524580471916124, 0.209755178075284, 0.855078564491123, 0.604500659974292, 0.449310077587143, 0.749021862866357, 0.981281586224213, 0.764989290386438, 0.254587037954479, 0.936133604031056, 0.761264442466199, 0.794549981830642, 0.77568340767175, 0.347362612141296, 0.858985813567415, 0.780902340309694, 0.994614759925753, 0.335018346086144, 0.240389818325639, 0.348824504064396, 0.949902991531417, 0.07659493223764, 0.446346551645547, 0.109486371278763, 0.672134774271399, 0.746278025908396, 0.686085518915206, 0.81868276395835, 0.518462428357452, 0.302804231178015, 0.654467743821442, 0.528389688814059, 0.880968261975795, 0.508492844179273, 0.806276791030541, 0.847784245153889, 0.89363248180598, 0.849843039410189, 0.305021597770974, 0.386938542593271, 0.401105006923899, 0.566703440388665, 0.276703527197242, 0.703231141902506, 0.8857726091519, 0.534183708485216, 0.459670649142936, 0.462384739192203, 0.808637630892918, 0.395194859243929, 0.582329931436107, 0.382292591501027, 0.810168859083205, 0.41757570952177, 0.855066460324451, 0.606522141257301, 0.480510390130803, 0.509133535437286, 0.169123193481937, 0.162566126557067, 0.564327326137573, 0.955694352742285, 0.600129895145074, 0.930818696273491, 0.632645287550986, 0.00747143849730492, 0.39807447581552, 0.503370095975697, 0.433914920780808, 0.543582562124357, 0.723954173270613, 0.389220718061551, 0.874785354128107, 0.222710429225117, 0.759848682209849, 0.2712711696513, 0.352411580272019, 0.643560910131782, 0.0118621326982975, 0.408553458750248, 0.0666154809296131, 0.34798714867793, 0.469685700489208, 0.333430896978825, 0.748068910092115, 0.505993503844365, 0.32437047502026, 0.392888332717121, 0.862115048803389, 0.459789904532954, 0.674139343667775, 0.659738802351058, 0.503044583369046, 0.248698593117297, 0.842161822831258, 0.977358027594164, 0.636723279720172, 0.257515648379922, 0.818328394554555, 0.905662501929328, 0.194341611349955, 0.630623556440696, 0.882296562427655, 0.249038366600871, 0.722022393252701, 0.0751412340905517, 0.115239507053047, 0.759771242504939, 0.746891697868705, 0.35774129233323, 0.744139926275238, 0.259275812422857, 0.733911121496931, 0.996483776019886, 0.287604777840897, 0.889630913501605, 0.056065117707476, 0.460969500476494, 0.762789081083611, 0.562876010080799, 0.36675727670081, 0.269354266580194, 0.633464424638078, 0.116752347676083, 0.402018199209124, 0.291649205377325, 0.993759607663378, 0.870429412461817, 0.0109584117308259, 0.516620509559289, 0.058529619127512, 0.539194704033434, 0.279678725404665, 0.886399830924347, 0.897120569134131, 0.219114861218259, 0.00531257479451597, 0.334628026233986, 0.552340265363455, 0.791186729213223, 0.444881453877315, 0.415322872111574, 0.234247926156968, 0.982129345880821, 0.232765212422237, 0.019486297853291, 0.165439231088385, 0.152036622166634, 0.192316161468625, 0.165611733682454, 0.154960548738018, 0.714299229206517, 0.15827712812461, 0.3430119024124, 0.845992579357699, 0.481281663291156, 0.427599991671741, 0.464043168351054, 0.0775466044433415, 0.901153416838497, 0.817228228785098, 0.105350905796513, 0.751990882446989, 0.155408964492381, 0.915659714955837, 0.27964580222033, 0.292602178873494, 0.000584352063015103, 0.735705047147349, 0.937328183092177, 0.701459996169433, 0.390517794294283, 0.296722705475986, 0.198973580962047, 0.898083057021722, 0.19166316324845, 0.430310661671683, 0.230338657274842, 0.92071468452923, 0.971666250377893, 0.392982916207984, 0.150212805718184, 0.544061968801543, 0.219475806690753, 0.258981329388916, 0.862225032178685, 0.594222264131531, 0.0182705922052264, 0.178085303166881, 0.005564323393628, 0.0601196431089193, 0.699003454297781, 0.79729702253826, 0.31125099491328, 0.990642984397709, 0.399245902430266, 0.735509477788582, 0.196145770372823, 0.764822692610323, 0.778937756782398, 0.65798480133526, 0.735163983656093, 0.914631095016375, 0.886479143286124, 0.808852086076513, 0.922011384507641, 0.905699030961841, 0.118413016432896, 0.193943365011364, 0.721142030321062, 0.358563068788499, 0.54070529085584, 0.290856401668862, 0.342471134150401, 0.253087423508987, 0.560667112004012, 0.295914644375443, 0.520887218182907, 0.610762525582686, 0.153022356098518, 0.380292525980622, 0.0625757323578, 0.392529719742015, 0.234147998038679, 0.367082656593993, 0.0376472868956625, 0.344180806307122, 0.807596567552537, 0.218906779307872, 0.621426794212312, 0.741973427124321, 0.518469255184755, 0.234369428595528, 0.831403268733993, 0.807660781778395, 0.920084153767675, 0.603042245609686, 0.0266597620211542, 0.188590502599254, 0.0384685697499663, 0.957985471002758, 0.386728142620996, 0.74463275866583, 0.439190730219707, 0.516854406800121, 0.961902637733147, 0.343100590631366, 0.693809157935902, 0.422649833373725, 0.10040626861155, 0.108267272822559, 0.45344763644971, 0.124986936338246, 0.111678895540535, 0.0182820337358862, 0.184191927546635, 0.234308956889436, 0.426860055886209, 0.739976201904938, 0.0961620665621012, 0.0699427735526115, 0.845586384646595, 0.459333651699126, 0.508311187615618, 0.217361043207347, 0.992972674779594, 0.242818285478279, 0.959414990153164, 0.195374505128711, 0.963141170330346, 0.661390480119735, 0.429945332696661, 0.652544667478651, 0.00469469255767763, 0.334498099051416, 0.78153076977469, 0.188013597158715, 0.277667501475662, 0.523717786651105, 0.0347726305481046, 0.142792386002839, 0.243872775696218, 0.0159939439035952, 0.528313405346125, 0.981783233582973, 0.108017327962443, 0.844816036289558, 0.26506075868383, 0.884178419364616, 0.343112249393016, 0.786370113957673, 0.0782496316824108, 0.913828439079225, 0.20626247022301, 0.564708990976214, 0.386258622165769, 0.703364262357354, 0.00714002852328122, 0.985413207439706, 0.565208162646741, 0.814480383414775, 0.187696422915906, 0.879016791237518, 0.410298397764564, 0.10832460783422, 0.584959441563115, 0.0678726462647319, 0.680252154823393, 0.0533722159452736, 0.673222502460703, 0.213033380452543, 0.979160605464131, 0.668721695896238, 0.970464298967272, 0.288495741551742, 0.0683744170237333, 0.944315909640864, 0.504796031396836, 0.872496198862791, 0.926198630593717, 0.423259238246828, 0.709695108234882, 0.198603755095974, 0.814357520313933, 0.862647085217759, 0.413917874684557, 0.805210701888427, 0.892802377929911, 0.128851839108393, 0.881327820010483, 0.90685446630232, 0.530864229425788, 0.663287753239274, 0.675794555339962, 0.699892327655107, 0.211146137909964, 0.181033862987533, 0.686201921431348, 0.965111561585218, 0.0623085116967559, 0.426675589289516, 0.224114496959373, 0.690576899098232, 0.394395803101361, 0.635655608726665, 0.283146815607324, 0.676657561445609, 0.735748139908537, 0.411254707025364, 0.180158596718684, 0.155278960941359, 0.284543743124232, 0.866343840956688, 0.162940016016364, 0.38363322801888, 0.704553934745491, 0.200594566762447, 0.158352338476107, 0.513984815217555, 0.149704845156521, 0.917596818180755, 0.767404436366633, 0.996247355593368, 0.250436670379713, 0.239972532028332, 0.2019773630891, 0.497079873457551, 0.330823141150177, 0.879641657229513, 0.614753192290664, 0.322863620007411, 0.175700078019872, 0.398685283493251, 0.74336817744188, 0.714613288873807, 0.858619545353577, 0.927810041001067, 0.50358924176544, 0.259426470845938, 0.500512540573254, 0.342983536422253, 0.717102170456201, 0.77399990754202, 0.00613413657993078, 0.516744387568906, 0.382351242471486, 0.0989413298666477, 0.197437224444002, 0.55132955010049, 0.293475500307977, 0.455413360148668, 0.0405781285371631, 0.523540630238131, 0.0393314347602427, 0.497186658671126, 0.0779311435762793, 0.0380314274225384, 0.542578055988997, 0.380058608250692, 0.401428107172251, 0.862857365282252, 0.766197687247768, 0.0970395030453801, 0.413646198809147, 0.212262289831415, 0.774683128576726, 0.500587242189795, 0.87344547174871, 0.871963174548, 0.251307619037107, 0.601806026650593, 0.633828054415062, 0.587904474698007, 0.66673385957256, 0.388582955114543, 0.564913093112409, 0.307152448222041, 0.333914724411443, 0.203989474335685, 0.0423696513753384, 0.610043172491714, 0.983209890080616, 0.637918004533276, 0.765093422494829, 0.0656715868972242, 0.544794486369938, 0.450357284164056, 0.793753660516813, 0.937112002400681, 0.944909701822326, 0.67561722593382, 0.327514302683994, 0.6805542006623, 0.426542797358707, 0.174600650090724, 0.283208105480298, 0.108165071113035, 0.875459138071164, 0.876375421183184, 0.547396377893165, 0.0532666987273842, 0.554535728413612, 0.0725639492738992, 0.941001974744722, 0.395221462706104, 0.181359730195254, 0.303733445471153, 0.082555569242686, 0.685309112071991, 0.842445831513032, 0.404968528542668, 0.385155471274629, 0.916806642198935, 0.298782144440338, 0.494931975612417, 0.698250718647614, 0.162563328165561, 0.370181670878083, 0.260566433658823, 0.847974545322359, 0.90246779914014, 0.367667697602883, 0.699669679393992, 0.33867095131427, 0.988107890123501, 0.240612608147785, 0.728050529025495, 0.758136505261064, 0.229334382107481, 0.225765229668468, 0.604724612785503, 0.349458849988878, 0.478547841310501, 0.842187750851735, 0.0715076222550124, 0.741184464190155, 0.384677454596385, 0.193787215044722, 0.191952694673091, 0.640578600112349, 0.672911992529407, 0.456495098304003, 0.363509436370805, 0.869370773434639, 0.0308212186209857, 0.646991209127009, 0.667658211896196, 0.0970854251645505, 0.811091996263713, 0.846536286175251, 0.93974539404735, 0.0883153544273227, 0.0738705871626735, 0.370294405845925, 0.87999126734212, 0.730221600737423, 0.562585371779278, 0.306292674038559, 0.89574665366672, 0.109515752876177, 0.913997885771096, 0.320526864379644, 0.679585514357314, 0.0958686047233641, 0.135616549290717, 0.755854186834767, 0.896362561034039, 0.67891710344702, 0.39063933189027, 0.464663684135303, 0.0236238881479949, 0.746424090582877, 0.226696040481329, 0.716831976315007, 0.873503393726423, 0.139701921259984, 0.770683886483312, 0.0251213668379933, 0.100544910645112, 0.16499840747565, 0.021626073634252, 0.383870366727933, 0.322163405129686, 0.687701671617106, 0.709791076369584, 0.0878219977021217, 0.25933146616444, 0.386758288368583, 0.334812856279314, 0.0386916790157557, 0.911875529913232, 0.312330664368346, 0.641552387969568, 0.732377366861328, 0.680730453925207, 0.922531319549307, 0.0756660064216703, 0.476072408258915, 0.417527189711109, 0.327152095735073, 0.786772493971512, 0.0754136152099818, 0.106491585960612, 0.900779436575249, 0.781994654797018, 0.80879493849352, 0.254124524071813, 0.271040274994448, 0.319067689822987, 0.644674311392009, 0.773905066540465, 0.192049509380013, 0.760934136807919, 0.56443163822405, 0.997339664725587, 0.847556181950495, 0.450546230655164, 0.0765253892168403, 0.334667721064761, 0.531505131395534, 0.07598831737414, 0.486315189162269, 0.364583496237174, 0.213001225842163, 0.97587401792407, 0.486405208241194, 0.396760666975752, 0.317067349329591, 0.742196901002899, 0.240943428594619, 0.0287311829160899, 0.361286589875817, 0.829878515563905, 0.0257482489105314, 0.505655366694555, 0.81921765813604, 0.157409723615274, 0.743526222882792, 0.394478993723169, 0.743776730261743, 0.604661862831563, 0.429558324394748, 0.212044054409489, 0.761004694737494, 0.442754289368168, 0.517849075142294, 0.4983081726823, 0.0862744920887053, 0.765467684715986, 0.11353668384254, 0.687514675781131, 0.742401836672798, 0.528225483838469, 0.808220444712788, 0.545000928686932, 0.473177272127941, 0.135143088409677, 0.0824297992512584, 0.925754774361849, 0.865836055250838, 0.912859762785956, 0.15683221956715, 0.22463344829157, 0.352179164532572, 0.27099243667908, 0.137257540831342, 0.155054670758545, 0.0412911782041192, 0.127664753934368, 0.637837508926168, 0.876253109192476, 0.825280916178599, 0.0433651674538851, 0.411273712292314, 0.607724698958918, 0.150724137900397, 0.337774134939536, 0.678095999639481, 0.700005379272625, 0.0555902610067278, 0.975707994773984, 0.465463019208983, 0.686116677243263, 0.500809692544863, 0.757523550884798, 0.349121950799599, 0.377634912962094, 0.874956225510687, 0.494438782334328, 0.00156154809519649, 0.0606482201255858, 0.0871747566852719, 0.176903548417613, 0.112732817884535, 0.106844818219543, 0.212203580187634, 0.602104633115232, 0.982098926324397, 0.958020927151665, 0.52484295819886, 0.21650868258439, 0.0969366603530943, 0.370156559394673, 0.49063866189681, 0.477566708810627, 0.300526724429801, 0.120949330041185, 0.332029955461621, 0.346197071950883, 0.294118235586211, 0.174762306502089, 0.889210475841537, 0.249768162146211, 0.460442641749978), actual = c(1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.528481996385381, 0.807959069032222, 0.903817334445193, 0.070748400175944, 0.730116463964805, 0.555241921450943, 0.467709330841899, 0.418055624933913, 0.557649738155305, 0.420845300424844, 0.339040268678218, 0.693250471958891, 0.457093569915742, 0.505144217284396, 0.547587531618774, 0.598478847648948, 0.296881868736818, 0.115530332783237, 0.0581681514158845, 0.664045916870236, 0.326550114899874, 0.63687631720677, 0.260697378311306, 0.173200140940025, 0.699063839158043, 0.455949924187735, 0.941026496002451, 0.190894427709281, 0.355923779075965, 0.967166731134057, 0.530433383537456, 0.836074429797009, 0.789286902407184, 0.577781406696886, 0.184111848007888, 0.586539134616032, 0.709689751034603, 0.205586256925017, 0.64396144868806, 0.0652496698312461, 0.689909253502265, 0.76279101963155, 0.311821756884456, 0.198646065313369, 0.0346376800443977, 0.092977192485705, 0.580226624850184, 0.385101945605129, 0.714936060830951, 0.0773666372988373, 0.276835387106985, 0.67648976482451, 0.999077868415043, 0.197900119237602, 0.314898017328233, 0.233935388503596, 0.0674829662311822, 0.805182506097481, 0.368806003825739, 0.555099170189351, 0.993842223659158, 0.351574928034097, 0.977990312501788, 0.65453294897452, 0.556356840999797, 0.8708986856509, 0.271879458101466, 0.0827265712432563, 0.567175392759964, 0.529976518591866, 0.852204209426418, 0.781490917317569, 0.321149331983179, 0.181372932158411, 0.603441105689853, 0.638107636710629, 0.259246941190213, 0.138017826480791, 0.562772156437859, 0.539612781489268, 0.747971323085949, 0.958428019890562, 0.756979633355513, 0.362952335970476, 0.731143607757986, 0.929620921844617, 0.766335988417268, 0.298714867560193, 0.321147248381749, 0.808895911090076, 0.269678497454152, 0.4811074398458, 0.275124560343102, 0.693372329697013, 0.896662327693775, 0.599168378161266, 0.589787106728181, 0.969116303604096, 0.0348991067148745, 0.0689746274147183, 0.360654707532376, 0.56513082748279, 0.724486828083172, 0.917282310081646, 0.0865758117288351, 0.284290517913178, 0.0344792264513671, 0.00465918541885912, 0.575524537824094, 0.749212307389826, 0.924312578048557, 0.209964377107099, 0.100580330938101, 0.512060916982591, 0.194621160859242, 0.851210683817044, 0.0349850070197135, 0.843305190559477, 0.938964570173994, 0.742452551377937, 0.448027594247833, 0.316972438246012, 0.130009717307985, 0.418806548928842, 0.288885629968718, 0.105919441441074, 0.891386889386922, 0.840667308773845, 0.758362844819203, 0.728970188181847, 0.195153183536604, 0.689747242024168, 0.380155808292329, 0.166916171088815, 0.0354702703189105, 0.352454880485311, 0.176253946730867, 0.0516146565787494, 0.925166003406048, 0.585085873026401, 0.970364579930902, 0.12839769711718, 0.811732708476484, 0.844846578780562, 0.938184359809384, 0.14009317336604, 0.873412028187886, 0.257045130245388, 0.0741231697611511, 0.391450310591608, 0.572081578429788, 0.594363787211478, 0.390144490869716, 0.0143954372033477, 0.123343374812976, 0.492648046463728, 0.231468664249405, 0.344828800763935, 0.497156508499756, 0.825889558997005, 0.269164523342624, 0.124740810366347, 0.0666266735643148, 0.421979592181742, 0.997576573630795, 0.646283690119162, 0.269560961285606, 0.131508370162919, 0.943214688682929, 0.344389805104584, 0.650515756802633, 0.73481949348934, 0.0865643601864576, 0.345343464286998, 0.915741528617218, 0.143264449667186, 0.304628293029964, 0.260401573264971, 0.417876345571131, 0.931943164439872, 0.657624418381602, 0.987474723951891, 0.84763674181886, 0.662479526363313, 0.483945549465716, 0.336780226556584, 0.285150237381458, 0.752674624556676, 0.260589656652883, 0.806776629528031, 0.534331508679315, 0.642301549203694, 0.506703118793666, 0.612095386488363, 0.927635312313214, 0.131909053772688, 0.346674832282588, 0.9927053747233, 0.564263462554663, 0.617808230686933, 0.620960805099458, 0.194753596791998, 0.893183501670137, 0.616663768189028, 0.0127462784294039, 0.0588164813816547, 0.418803479056805, 0.0160511201247573, 0.339141817297786, 0.987386756576598, 0.795672939624637, 0.729551359545439, 0.396466020494699, 0.981387717183679, 0.854230768978596, 0.0674412548542023, 0.946171335643157, 0.836607294622809, 0.536593455821276, 0.0577064615208656, 0.99059461732395, 0.358691611792892, 0.11964240600355, 0.442519790958613, 0.470265452982858, 0.461215787567198, 0.997018414083868, 0.399261664832011, 0.180061427177861, 0.357583319069818, 0.832845209399238, 0.365421219961718, 0.592634614324197, 0.814620892051607, 0.92762353271246, 0.235381563194096, 0.724469902459532, 0.216674017952755, 0.651129858568311, 0.679482539417222, 0.507366199046373, 0.361216455232352, 0.617046060273424, 0.247212874237448, 0.0285942552145571, 0.481495009968057, 0.968118528835475, 0.164546742569655, 0.898128347005695, 0.86613807361573, 0.173753112088889, 0.112783371703699, 0.219233319628984, 0.948291974607855, 0.16655381815508, 0.273007611976936, 0.965292819775641, 0.964298280421644, 0.507970093283802, 0.272183356806636, 0.818195171887055, 0.594812966650352, 0.254150317516178, 0.951314905891195, 0.854156454792246, 0.821485107997432, 0.538059096084908, 0.685848672874272, 0.703260098118335, 0.392883520340547, 0.989116691751406, 0.952687624143437, 0.635572172701359, 0.539232710609213, 0.0522100899834186, 0.899861812591553, 0.686410293215886, 0.552204348379746, 0.251238271361217, 0.0638116491027176, 0.451716942014173, 0.740026477025822, 0.898558968678117, 0.913079443387687, 0.875375430798158, 0.3488948661834, 0.384623334975913, 0.152732673333958, 0.639068516669795, 0.696175633231178, 0.262748968787491, 0.099222068907693, 0.522890098858625, 0.498930721310899, 0.665113838389516, 0.478523787809536, 0.159082373138517, 0.794907255098224, 0.659090543864295, 0.597185176797211, 0.733568460447714, 0.799326161388308, 0.982544230530038, 0.64332280959934, 0.701463317032903, 0.718205468030646, 0.94427022524178, 0.340512045426294, 0.0858992517460138, 0.0511797065846622, 0.300841105636209, 0.579267028253525, 0.4614338260144, 0.560909246327356, 0.583246297203004, 0.120867667719722, 0.77693225373514, 0.538670205743983, 0.366170912981033, 0.210378310177475, 0.73738218517974, 0.0184564334340394, 0.849184709368274, 0.290765641024336, 0.977864390704781, 0.802013401407748, 0.410641090245917, 0.934025081573054, 0.0616843195166439, 0.936280856607482, 0.671588782221079, 0.868398105492815, 0.130813775584102, 0.627090345602483, 0.570468441117555, 0.158922819187865, 0.882689823396504, 0.308742340654135, 0.267414804315194, 0.0879052244126797, 0.87412239680998, 0.354247559793293, 0.121085795806721, 0.262854824773967, 0.606672234833241, 0.144903888460249, 0.246580396546051, 0.827385565266013, 0.803160631330684, 0.984847610583529, 0.40590562694706, 0.263025102438405, 0.699838487431407, 0.903666488127783, 0.58052774053067, 0.605303007178009, 0.24851634632796, 0.83283160906285, 0.305281687295064, 0.0769365937449038, 0.794272580184042, 0.463737402809784, 0.0220352811738849, 0.563733468297869, 0.159281021682546, 0.571856132475659, 0.736725184833631, 0.50499132880941, 0.340074371779338, 0.972165974089876, 0.88096443680115, 0.683428349206224, 0.859299314441159, 0.199042240157723, 0.942821150878444, 0.326272519305348, 0.739583930699155, 0.183748586801812, 0.0142797185108066, 0.669325655559078, 0.0477788264397532, 0.375446947757155, 0.757403412368149, 0.621094383299351, 0.546278897905722, 0.813071232987568, 0.706036269897595, 0.513242549961433, 0.98394008516334, 0.699561478104442, 0.728356763953343, 0.272495932178572, 0.279492660425603, 0.435240915743634, 0.601907465374097, 0.431423247558996, 0.249980162363499, 0.166160169057548, 0.486870009219274, 0.784375749528408, 0.412136890226975, 0.696520811179653, 0.13370167883113, 0.540129100671038, 0.706647705286741, 0.548771896399558, 0.209760834695771, 0.707774813752621, 0.607526575680822, 0.390338702825829, 0.301412577042356, 0.729383022291586, 0.274796851444989, 0.497261880431324, 0.535506567917764, 0.48641106672585, 0.899190235882998, 0.439241724321619, 0.906088372692466, 0.898970507085323, 0.387536184629425, 0.836972278775647, 0.653153938008472, 0.724395268829539, 0.0372396141756326, 0.574364696396515, 0.139132922515273, 0.525713765295222, 0.528754697181284, 0.621837251354009, 0.0536450061481446, 0.052425317466259, 0.386411367217079, 0.991918667918071, 0.663274591555819, 0.935333397006616, 0.621107342187315, 0.451986639993265, 0.0373670484405011, 0.438171692425385, 0.896400126395747, 0.435155495768413, 0.355112865334377, 0.834905503550544, 0.218002972193062, 0.0661721171345562, 0.987850214121863, 0.82620060024783, 0.625400797231123, 0.446916072862223, 0.434457282768562, 0.0849186780396849, 0.355894975131378, 0.522662046132609, 0.631569784833118, 0.829776884289458, 0.783540095901117, 0.374668423784897, 0.160819847369567, 0.420835739467293, 0.920247954549268, 0.177483315113932, 0.387005835771561, 0.744501533685252, 0.724438471719623, 0.835743084084243, 0.656720628496259, 0.0977372573688626, 0.292779413284734, 0.539836474694312, 0.324551826808602, 0.172386551043019, 0.0332431541755795, 0.00199513928964734, 0.528990999329835, 0.457058107713237, 0.757788300281391, 0.0547134093940258, 0.225783334579319, 0.289119731169194, 0.615098709007725, 0.445070444839075, 0.503026368329301, 0.451796280452982, 0.517347533488646, 0.940652609104291, 0.415750172222033, 0.602311462163925, 0.6190604288131, 0.981379674514756, 0.456709889462218, 0.395285139326006, 0.367916436400265, 0.475033103488386, 0.763916647993028, 0.0810136501677334, 0.219272113870829, 0.385264018317685, 0.462551299715415, 0.790382320294157, 0.113526738714427, 0.422083019511774, 0.00770472339354455, 0.875768771162257, 0.73480702470988, 0.555894070537761, 0.18501696921885, 0.431507617002353, 0.581123386975378, 0.00503316638059914, 0.0514991516247392, 0.344889465486631, 0.715779592515901, 0.503344624536112, 0.849286751821637, 0.00360240694135427, 0.297763268230483, 0.356377934571356, 0.992008289089426, 0.40128744370304, 0.0541539425030351, 0.213158222381026, 0.643426266964525, 0.686281937174499, 0.765402095392346, 0.682725758291781, 0.954367475351319, 0.0751550446730107, 0.851090541342273, 0.0345490877516568, 0.507343441015109, 0.526432055747136, 0.233679506229237, 0.637172074755654, 0.847704177722335, 0.27444658242166, 0.552800090052187, 0.459106743335724, 0.792510622413829, 0.310398207977414, 0.970931437797844, 0.667795114917681, 0.768839187454432, 0.0288040891755372, 0.81836248585023, 0.450391419231892, 0.605883352924138, 0.973758208798245, 0.462642560945824, 0.937289715511724, 0.763414961285889, 0.517517621163279, 0.0473668382037431, 0.161966381594539, 0.317418941296637, 0.755724526476115, 0.527993470663205, 0.467289416585118, 0.319424797547981, 0.663957266602665, 0.473330486798659, 0.808358542155474, 0.655617988901213, 0.0594322311226279, 0.354287083493546, 0.695149853825569, 0.456145816249773, 0.00836474122479558, 0.792187514016405, 0.822012309683487, 0.323487739311531, 0.329644215526059, 0.534527273150161, 0.152374696685001, 0.0740629923529923, 0.0736252195201814, 0.750785660231486, 0.400640359148383, 0.617106728954241, 0.920617380412295, 0.518152377801016, 0.292964746477082, 0.279068533331156, 0.246186883421615, 0.580707992427051, 0.467815554467961, 0.198413035832345, 0.03625791775994, 0.69063834589906, 0.427202934166417, 0.521398641867563, 0.420238496968523, 0.982444952707738, 0.486950370715931, 0.337845474481583, 0.429137017810717, 0.307834060164168, 0.0373764999676496, 0.0140885394066572, 0.506355725694448, 0.879936665762216, 0.242766063194722, 0.873603821964934, 0.496916697593406, 0.360071774339303, 0.822621149942279, 0.961106398375705, 0.0734795513562858, 0.0548277997877449, 0.268180353799835, 0.0745022026821971, 0.0700232579838485, 0.940504079917446, 0.95422689570114, 0.4627542917151, 0.517958483658731, 0.402844990836456, 0.608978370204568, 0.373260327382013, 0.371678961906582, 0.65314903925173, 0.439734450308606, 0.908539939904585, 0.962867312831804, 0.250540753128007, 0.756642444990575, 0.554813072783872, 0.565749713685364, 0.591857847757638, 0.256790463114157, 0.130792480660602, 0.0634554512798786, 0.463151479372755, 0.925059852190316, 0.559663094813004, 0.511598238488659, 0.960907984059304, 0.860726810758933, 0.796630742494017, 0.0304535815957934, 0.572753882268444, 0.132387364050373, 0.879771334351972, 0.750330629525706, 0.687418596353382, 0.592024880461395, 0.656891656806692, 0.0156725549604744, 0.0361483125016093, 0.370888395700604, 0.696402563713491, 0.838344450341538, 0.285816353978589, 0.459977123653516, 0.848962335614488, 0.182480458868667, 0.845469697611406, 0.724304353119805, 0.646331163588911, 0.577052594861016, 0.945049892645329, 0.224971173796803, 0.0712965768761933, 0.00404419447295368, 0.286527439486235, 0.112707666819915, 0.534581645857543, 0.0493960408493876, 0.665854629827663, 0.148829962592572, 0.358284644782543, 0.94652570062317, 0.354638931341469, 0.316509008873254, 0.894264587201178, 0.117945456877351, 0.345317404018715, 0.548402379732579, 0.39473447878845, 0.224646344315261, 0.636820293264464, 0.586955650476739, 0.372262064367533, 0.0603328957222402, 0.400667714886367, 0.46840289211832, 0.728086337679997, 0.970358862541616, 0.945830419659615, 0.0123524381779134, 0.910284626763314, 0.434126541484147, 0.0531262073200196, 0.376741889165714, 0.626069082412869, 0.928030414739624, 0.278580745216459, 0.408906182274222, 0.205165862804279, 0.0671680578961968, 0.0368553404696286, 0.185864624800161, 0.330813102424145, 0.802694737911224, 0.0156051574740559, 0.0403179340064526, 0.0158722624182701, 0.365299008786678, 0.646197522990406, 0.258351013297215, 0.144691111985594, 0.413247650722042, 0.124106442322955, 0.61054613487795, 0.243994207354262, 0.103892500279471, 0.362421848811209, 0.116794318659231, 0.166020647855476, 0.348754306323826, 0.725182248046622, 0.346926850499585, 0.237099595600739, 0.188754630740732, 0.974093064432964, 0.363050628453493, 0.112617874983698, 0.183090443955734, 0.618846267927438, 0.284830923890695, 0.490872607333586, 0.38722427142784, 0.770469878334552, 0.252357745775953, 0.797389894956723, 0.394014991587028, 0.856521511916071, 0.0817706899251789, 0.45894314115867, 0.679392439313233, 0.294777454808354, 0.957759570097551, 0.817355754319578, 0.70251851156354, 0.846668637124822, 0.00407429388724267, 0.0675692991353571, 0.764082658104599, 0.790162381483242, 0.921359336702153, 0.21837101620622, 0.923047664109617, 0.701567703625187, 0.410461839055642, 0.61210622661747, 0.169735154369846, 0.628193634096533, 0.515655611641705, 0.82070983457379, 0.965019499184564, 0.810434156795964, 0.86111725284718, 0.282129645114765, 0.479591894196346, 0.495923473965377, 0.122900949558243, 0.595882727298886, 0.0704041889403015, 0.821110866731033, 0.0451976722106338, 0.974179835058749, 0.646317730890587, 0.685616694623604, 0.0308045169804245, 0.512440622551367, 0.0296770553104579, 0.743362500797957, 0.615327525185421, 0.907130474457517, 0.595343276159838, 0.947764480020851, 0.462541721528396, 0.481826942879707, 0.361184846377, 0.609576666494831, 0.691668525803834, 0.946826770436019, 0.116233949782327, 0.609020828735083, 0.0113213816657662, 0.650411701295525, 0.855668465839699, 0.739463102305308, 0.505752591881901, 0.421805232996121, 0.428164535202086, 0.0553618213161826, 0.758214467205107, 0.918587415711954, 0.00889152474701405, 0.791250419802964, 0.538014979334548, 0.461712806485593, 0.430521161993966, 0.469094663392752, 0.0498067217413336, 0.326664407271892, 0.7688168133609, 0.959908740362152, 0.907811571611091, 0.596614338224754, 0.990085674216971, 0.0611196123063564, 0.0645097508095205, 0.70971883716993, 0.664849930210039, 0.466316593345255, 0.746796790510416, 0.613057716051117, 0.551962602650747, 0.0798446408007294, 0.244456457672641, 0.58932485524565, 0.218328855698928, 0.494892133865505, 0.979721528477967, 0.563054481055588, 0.893640997353941, 0.122509701410308, 0.790558795444667, 0.62246615299955, 0.727469083620235, 0.581095190485939, 0.918064965400845, 0.294273298000917, 0.829454192658886, 0.913009064970538, 0.558595017530024, 0.77247441932559, 0.395555397728458, 0.161286165704951, 0.136858770856634, 0.553911161376163, 0.749645488802344, 0.118598052766174, 0.698255649302155, 0.852246051421389, 0.754299910506234, 0.410570206353441, 0.0313945384696126, 0.339270626194775, 0.464827601332217, 0.844192409189418, 0.574713981244713, 0.29452029732056, 0.67835721350275, 0.935339683201164, 0.602499497123063, 0.801350099034607, 0.436757256509736, 0.358843728667125, 0.707844663178548, 0.695872667478397, 0.500619941856712, 0.504982257261872, 0.358208248158917, 0.853413652861491, 0.345079242717475, 0.27484345715493, 0.798459920566529, 0.28745069890283, 0.954065732657909, 0.643085626419634, 0.631188893690705, 0.285416111117229, 0.276486862218007, 0.437172936275601, 0.296823097160086, 0.0641887900419533, 0.851037485757843, 0.114129253895953, 0.913161390228197, 0.314947833772749, 0.288166906218976, 0.166173302102834, 0.710797481471673, 0.889323245501146, 0.282021086197346, 0.583684789016843, 0.415605513844639, 0.856672680471092, 0.0303653236478567, 0.226231356617063, 0.707729078130797, 0.644099840661511, 0.482811631402001, 0.470147185493261, 0.329199898522347, 0.259819883853197, 0.528913276270032, 0.503006311366335, 0.965725186513737, 0.10873135807924, 0.861780864186585, 0.280466878088191, 0.55329452524893, 0.678465203382075, 0.554732667282224, 0.51664353068918, 0.0166643091943115, 0.0998128121718764, 0.323732550954446, 0.685578681062907, 0.841113447910175, 0.91969548817724, 0.0384937720373273, 0.073734253179282, 0.0217860892880708, 0.892599257174879, 0.706559492507949, 0.177754151402041, 0.0159585487563163, 0.628217201447114, 0.998982629738748, 0.883690622868016, 0.777766633080319, 0.218542558373883, 0.696818528696895, 0.962942339479923, 0.584818233270198, 0.214557176688686, 0.145970085635781, 0.47035562293604, 0.421773487702012, 0.738742815330625, 0.561721635283902, 0.194600182585418, 0.942699252162129, 0.326924451859668, 0.672171885380521, 0.840467673959211, 0.464135623304173, 0.115165996365249, 0.525035084923729, 0.0378087048884481, 0.947207000805065, 0.00713341101072729, 0.351501741912216, 0.400859850458801, 0.634041983867064, 0.0482262286823243, 0.450017574941739, 0.106471188133582, 0.975921203847975, 0.316379110794514, 0.291764029767364, 0.30309717124328, 0.660504850558937, 0.202937134541571, 0.774844042491168, 0.948973097838461, 0.0425231161061674, 0.873568112030625, 0.245458311866969, 0.210140047594905, 0.823556409450248, 0.623728445498273, 0.195552315330133, 0.332280345493928, 0.76784196542576, 0.0857810261659324, 0.920295987976715, 0.166876713512465, 0.903613059083, 0.937899186275899, 0.566770744509995, 0.704143356764689, 0.599460999714211, 0.213102775393054, 0.457446160493419, 0.510713518830016, 0.290419338271022, 0.234288403997198, 0.0332027731928974, 0.0294931773096323, 0.564278802601621, 0.807757091475651, 0.50381187023595, 0.675825378391892, 0.188498719595373, 0.490341646596789, 0.0607521294150501, 0.226024371106178, 0.651421951595694, 0.0166270439513028, 0.246641620760784, 0.101150738773867, 0.145437805913389, 0.0117838245350868, 0.522963834460825), actual = c(0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"