Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.1375670        0.1375670 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5128538        0.5128538        0.5087177 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.246615683892742, 0.946816301438957, 0.229408192448318, 0.766201118938625, 0.830970703624189, 0.120656710118055, 0.258822502102703, 0.215751081006601, 0.0135505511425436, 0.669089215574786, 0.847625361056998, 0.489484137389809, 0.559112548595294, 0.656120125669986, 0.0216670522931963, 0.543964532902464, 0.200914046959952, 0.972045804141089, 0.553793512051925, 0.217564870137721, 0.441019848920405, 0.396381603786722, 0.603949395241216, 0.486353120533749, 0.217788794543594, 0.281040070112795, 0.760375448735431, 0.15509290038608, 0.194104809081182, 0.406481227604672, 0.626714649610221, 0.484635845990852, 0.324221260612831, 0.834819484967738, 0.467001246288419, 0.761280036531389, 0.949139751493931, 0.489172058878466, 0.570171338273212, 0.352411844534799, 0.733355895383283, 0.981360295088962, 0.559380836552009, 0.967651635874063, 0.829459729604423, 0.129407193744555, 0.516888160957024, 0.859841553727165, 0.000705517595633864, 0.0699888768140227, 0.167584037873894, 0.321334285428748, 0.905414135428146, 0.0837695824448019, 0.273181698517874, 0.14035533927381, 0.0169485621154308, 0.0921421546954662, 0.17393907206133, 0.0796821673866361, 0.111620200565085, 0.956360715441406, 0.523835632484406, 0.449578269151971, 0.848686208948493, 0.428955462761223, 0.459600247209892, 0.469111895188689, 0.755758679239079, 0.173574468586594, 0.646953622577712, 0.0984499966725707, 0.558383990311995, 0.48258721944876, 0.0717773581854999, 0.529970360221341, 0.537754222983494, 0.206600651377812, 0.408567079342902, 0.476553798420355, 0.769926203181967, 0.978580051334575, 0.832774905487895, 0.872565515106544, 0.9229396879673, 0.729259787593037, 0.344174882629886, 0.919417712837458, 0.769111308269203, 0.42518436932005, 0.748965692240745, 0.32517713541165, 0.766091909492388, 0.344456509454176, 0.628183613996953, 0.916846036445349, 0.709407242247835, 0.240283318795264, 0.139406931120902, 0.848470077151433, 0.754643256543204, 0.90494483220391, 0.118182346923277, 0.930724167497829, 0.220663005951792, 0.952588711865246, 0.738387247547507, 0.219810759648681, 0.673402077285573, 0.478717490565032, 0.675975817721337, 0.808845753083006, 0.292300135362893, 0.0146618885919452, 0.514060520567, 0.932426995132118, 0.0901140903588384, 0.887542938115075, 0.177754203788936, 0.246730366256088, 0.330064893467352, 0.555089317495003, 0.10267759533599, 0.378926026867703, 0.0500058885663748, 0.138130936771631, 0.413064732449129, 0.470108862966299, 0.643831316614524, 0.793592966161668, 0.407038730802014, 0.0640512297395617, 0.881093975389376, 0.0155554013326764, 0.781436670105904, 0.827370064333081, 0.294239745242521, 0.400318074971437, 0.131270048674196, 0.0389167035464197, 0.736606115009636, 0.0664863295387477, 0.021189181599766, 0.231857951031998, 0.678670326480642, 0.151420298963785, 0.673433068441227, 0.806356884539127, 0.72169952839613, 0.818127340404317, 0.601030993042514, 0.317907794145867, 0.744435859145597, 0.899129411205649, 0.289052870124578, 0.432116995332763, 0.507390617625788, 0.265534417238086, 0.229039745870978, 0.243886320851743, 0.134229108924046, 0.283181373495609, 0.762277541449293, 0.63865911681205, 0.702890371438116, 0.216997919837013, 0.743307195138186, 0.696980334352702, 0.171278207330033, 0.0450208708643913, 0.839402279118076, 0.289453190984204, 0.775528379483148, 0.664930084953085, 0.978179301600903, 0.636081017786637, 0.593377200188115, 0.831595808267593, 0.656870587030426, 0.0864202461671084, 0.919726345920935, 0.90710542188026, 0.55162546550855, 0.651640432421118, 0.628036931855604, 0.754218635847792, 0.307416890515015, 0.971657248679549, 0.503930080682039, 0.58209941140376, 0.945630554808304, 0.826277580112219, 0.683025243226439, 0.896708273328841, 0.219437773572281, 0.0341775701381266, 0.187483714427799, 0.0949841793626547, 0.455160148208961, 0.44630598789081, 0.818653298076242, 0.0543658984825015, 0.75250103441067, 0.0246424332726747, 0.204556811368093, 0.572390607791021, 0.50335785606876, 0.358030008152127, 0.743086414644495, 0.0486382127273828, 0.672207685885951, 0.228892066283152, 0.675574619323015, 0.97114711557515, 0.66952248592861, 0.767456582048908, 0.736146945506334, 0.171864006202668, 0.215213061543182, 0.389593292027712, 0.187604373320937, 0.48803822719492, 0.71958315349184, 0.60971907293424, 0.0203768343199044, 0.562677692621946, 0.225422683404759, 0.932499197777361, 0.318289724411443, 0.729925496969372, 0.560962620656937, 0.0649985261261463, 0.405343551887199, 0.829303464386612, 0.489038577768952, 0.639837148832157, 0.941309008281678, 0.845242458162829, 0.552388799842447, 0.534780914662406, 0.720738533651456, 0.210054684430361, 0.683468265691772, 0.119985431665555, 0.575229679932818, 0.481598472921178, 0.9667988861911, 0.703607600647956, 0.275629986310378, 0.714101724792272, 0.602043426129967, 0.460501773515716, 0.883385261520743, 0.563000355381519, 0.54452312993817, 0.678522304166108, 0.31913033593446, 0.200652964645997, 0.10381573997438, 0.969593998976052, 0.387113908305764, 0.894660489633679, 0.604087168117985, 0.422862107399851, 0.180834745755419, 0.859795835334808, 0.927142129978165, 0.979381469776854, 0.913524028146639, 0.885439937002957, 0.77457879902795, 0.40765589592047, 0.621956378221512, 0.481966765830293, 0.76919582625851, 0.197111097630113, 0.362685377476737, 0.105587954400107, 0.804057647939771, 0.395601527299732, 0.380173448007554, 0.65500974142924, 0.565361538901925, 0.316142281750217, 0.666143548674881, 0.481377165298909, 0.862638769671321, 0.579094748245552, 0.686336326180026, 0.354528509546071, 0.567707840353251, 0.132417219923809, 0.726891640108079, 0.403245789464563, 0.189543574349955, 0.205979620339349, 0.37702644825913, 0.481104749953374, 0.725463701644912, 0.35300115426071, 0.192974789068103, 0.821740755811334, 0.400178692536429, 0.525947663234547, 0.800792125519365, 0.0650470382533967, 0.444670256925747, 0.676286607980728, 0.508744056103751, 0.976510358043015, 0.232855246169493, 0.571297359652817, 0.0813545249402523, 0.73220121790655, 0.0239862946327776, 0.70550426049158, 0.611685378476977, 0.869783623376861, 0.280905685620382, 0.675589496269822, 0.240896858042106, 0.0670451193582267, 0.324112405069172, 0.743575089378282, 0.67924445332028, 0.430275857215747, 0.196008061990142, 0.442360322223976, 0.25894562411122, 0.0968838669359684, 0.376081004040316, 0.988693806575611, 0.175111602526158, 0.473455678438768, 0.275448700180277, 0.511548259994015, 0.242084073834121, 0.0127809362020344, 0.691745787393302, 0.341509661404416, 0.298061460955068, 0.970132356975228, 0.442390208598226, 0.691778533160686, 0.797180178808048, 0.0868211218621582, 0.0131377768702805, 0.519906905246899, 0.940774891292676, 0.0017773995641619, 0.520738299936056, 0.887126542860642, 0.471666090423241, 0.146975020412356, 0.638039117679, 0.136883296305314, 0.325333627639338, 0.167470093816519, 0.497510875575244, 0.0923722384031862, 0.430639667902142, 0.624497502110898, 0.45475639984943, 0.21013321285136, 0.67757199332118, 0.896282253786922, 0.0784374978393316, 0.174535041907802, 0.792056929087266, 0.387717663776129, 0.784206765936688, 0.167207152815536, 0.0112780712079257, 0.98087531956844, 0.404154952149838, 0.334713550517336, 0.259915580274537, 0.381442919140682, 0.019518778193742, 0.569120991975069, 0.707568804966286, 0.626070290803909, 0.254982570186257, 0.688192604575306, 0.959077131235972, 0.535950401099399, 0.136295088334009, 0.631436421535909, 0.0499011329375207, 0.510819159680977, 0.893109770026058, 0.357750267256051, 0.0455092734191567, 0.865817993879318, 0.304042162373662, 0.643202649662271, 0.907400493510067, 0.854766631731763, 0.159592054318637, 0.724489784799516, 0.457409806549549, 0.790884351823479, 0.339720507385209, 0.9355199332349, 0.978319918736815, 0.0299047965090722, 0.435092620784417, 0.0801003479864448, 0.0933101994451135, 0.586227986263111, 0.439118139911443, 0.92082802997902, 0.0181268306914717, 0.0246023966465145, 0.138283746084198, 0.129066031426191, 0.288967077387497, 0.774476285092533, 0.475985529599711, 0.354407137725502, 0.57479186472483, 0.296442851191387, 0.824309565359727, 0.619052660651505, 0.365487636998296, 0.899741241242737, 0.512584320269525, 0.67553069931455, 0.594184340676293, 0.142572363372892, 0.566293480107561, 0.195424389559776, 0.318732786225155, 0.48835720284842, 0.352758966153488, 0.698340041330084, 0.649095678934827, 0.180057963822037, 0.0164104432333261, 0.0453398434910923, 0.0799369129817933, 0.978677664883435, 0.257649676175788, 0.142166824778542, 0.266367185860872, 0.0259005825500935, 0.0933940159156919, 0.379989277105778, 0.64469846105203, 0.012870789738372, 0.699015133781359, 0.70270067662932, 0.338243163656443, 0.846924211829901, 0.7273036937695, 0.118746033636853, 0.228189918911085, 0.598547710571438, 0.141636901302263, 0.234259767225012, 0.947565686423331, 0.315297358902171, 0.959213066613302, 0.779634342063218, 0.254540713038296, 0.336810069158673, 0.885224614292383, 0.39377993857488, 0.848597201751545, 0.1611752146855, 0.728569200262427, 0.325955532258376, 0.709531484404579, 0.996598422294483, 0.0721528031863272, 0.685649656457826, 0.786932277260348, 0.587421196978539, 0.538984769256786, 0.543758532498032, 0.0361898299306631, 0.371848908718675, 0.750781321898103, 0.563317416235805, 0.319930274505168, 0.64515887433663, 0.880251100054011, 0.0399400775786489, 0.666053058346733, 0.00727060064673424, 0.339400109136477, 0.630272233393043, 0.54446553462185, 0.920326644321904, 0.591571221360937, 0.737504315096885, 0.341913491021842, 0.429833317641169, 0.489496644819155, 0.760187080129981, 0.119934817543253, 0.990730710793287, 0.762608463177457, 0.00509448000229895, 0.120881574228406, 0.22363377478905, 0.722116025397554, 0.788395596202463, 0.937063102144748, 0.0943620640318841, 0.748714386019856, 0.58006510254927, 0.556446892209351, 0.923344197450206, 0.407342195743695, 0.493009383324534, 0.683022054610774, 0.408387382747605, 0.374567972728983, 0.299476739251986, 0.940050170989707, 0.715475864242762, 0.560488890390843, 0.192497075069696, 0.53540265886113, 0.515183286508545, 0.0645890454761684, 0.925693171098828, 0.94897991232574, 0.00643769628368318, 0.9352445024997, 0.900245689786971, 0.256311930483207, 0.928403763333336, 0.65010604634881, 0.40153742628172, 0.078061148757115, 0.965897885849699, 0.582644730107859, 0.694334390573204, 0.414944068063051, 0.670983912888914, 0.0346500491723418, 0.0686012592632324, 0.770223141647875, 0.446980896871537, 0.868415936129168, 0.438670411007479, 0.502566708484665, 0.342971756821498, 0.825930217513815, 0.0332187307067215, 0.204158396460116, 0.759038808522746, 0.255158412270248, 0.484154570614919, 0.629419020842761, 0.818602093029767, 0.810546318301931, 0.0341641327831894, 0.238479931373149, 0.903452523052692, 0.720050151459873, 0.323279137955979, 0.953022163826972, 0.0923725222237408, 0.572530787205324, 0.535340582719073, 0.318762794835493, 0.471668178215623, 0.420659095281735, 0.327462970046327, 0.088934809435159, 0.756683375220746, 0.652648370712996, 0.0886569048743695, 0.176283810287714, 0.266152427066118, 0.576164987171069, 0.42936357925646, 0.0399415383581072, 0.136748205637559, 0.131740193348378, 0.590279759839177, 0.0751415607519448, 0.991109519032761, 0.848464037291706, 0.300677662948146, 0.967056494671851, 0.48176472261548, 0.364242832176387, 0.0471765510737896, 0.46907062176615, 0.266991924727336, 0.522522874874994, 0.478053128579631, 0.299209943041205, 0.554533818503842, 0.447408564621583, 0.889568131417036, 0.690050537930802, 0.205578913446516, 0.559313998091966, 0.717040226794779, 0.0474636382423341, 0.0674890212249011, 0.426584831438959, 0.746627007145435, 0.68045395379886, 0.32166481600143, 0.586443678941578, 0.544801358832046, 0.758816461544484, 0.222712869057432, 0.473357650218531, 0.839082324644551, 0.270974088693038, 0.250109701184556, 0.185361379059032, 0.107525552157313, 0.479218857362866, 0.73210180667229, 0.0805991040542722, 0.821472256677225, 0.389024700969458, 0.153658214025199, 0.0877067623659968, 0.713435092708096, 0.264113062061369, 0.8639485782478, 0.404453133000061, 0.905292144045234, 0.99073189473711, 0.0109472121112049, 0.121398336719722, 0.51215813937597, 0.138767366763204, 0.132423091912642, 0.0429262092802674, 0.347370628500357, 0.790241501992568, 0.901841029059142, 0.197567462921143, 0.0126881813630462, 0.154533569933847, 0.863071927800775, 0.650584275368601, 0.0762060151901096, 0.706428823061287, 0.407356917159632, 0.637753236573189, 0.186509317485616, 0.309742243029177, 0.771652382100001, 0.517100870376453, 0.857127680676058, 0.0232587088830769, 0.0643983178306371, 0.378869004081935, 0.140901036560535, 0.851849421858788, 0.51552904304117, 0.283829953987151, 0.924905152060091, 0.823531885631382, 0.841530021978542, 0.659604815300554, 0.762891534715891, 0.235304677160457, 0.861826135311276, 0.798942446708679, 0.603770208079368, 0.0802319413051009, 0.380672418279573, 0.759865406202152, 0.674606934888288, 0.12212822213769, 0.159276872174814, 0.493386416463181, 0.610343882348388, 0.494982805335894, 0.940324211493134, 0.508753359084949, 0.14209747617133, 0.817445507505909, 0.667227284051478, 0.67417546059005, 0.578363543376327, 0.817706889705732, 0.146113712107763, 0.478721038205549, 0.528134013293311, 0.109667624114081, 0.593048819573596, 0.482189294183627, 0.475796358427033, 0.389823748730123, 0.460000685183331, 0.946489078458399, 0.22080619004555, 0.475689558777958, 0.300973534118384, 0.161925176391378, 0.41130140167661, 0.149078517686576, 0.721957652829587, 0.150105261942372, 0.779284812277183, 0.670073850546032, 0.44070117152296, 0.460547187831253, 0.529841324547306, 0.502898021833971, 0.35630417871289, 0.194006219506264, 0.456998790148646, 0.182722053257748, 0.815987586276606, 0.0912037952803075, 0.070242895046249, 0.29809513525106, 0.684750229353085, 0.511247587855905, 0.769338755169883, 0.665389581816271, 0.0243429106194526, 0.281936932122335, 0.395026371115819, 0.770244773477316, 0.672353739850223, 0.095292836194858, 0.697424670448527, 0.197455454850569, 0.0837654555216432, 0.101484008599073, 0.915065402165055, 0.501629489473999, 0.480667297262698, 0.763367761624977, 0.969061643118039, 0.277023273520172, 0.910914656473324, 0.525518444133922, 0.503899143310264, 0.411990261403844, 0.581897518597543, 0.621511316625401, 0.248589370632544, 0.476534681394696, 0.712149309692904, 0.63712560525164, 0.197234201710671, 0.0297776297666132, 0.00977900833822787, 0.335488478187472, 0.0831968241836876, 0.123786080395803, 0.792337807128206, 0.241366766160354, 0.482005723519251, 0.32728425017558, 0.877861597575247, 0.346092190593481, 0.586926299612969, 0.537931899074465, 0.307349788956344, 0.499244409846142, 0.345764491241425, 0.875062564853579, 0.11717783450149, 0.44917703513056, 0.560208319919184, 0.762231854489073, 0.418439088622108, 0.0882888995110989, 0.0755261066369712, 0.52206704299897, 0.792411643313244, 0.233889372320846, 0.453203267883509, 0.625265763839707, 0.0564667128492147, 0.727371755288914, 0.22840539435856, 0.777371184667572, 0.61360823479481, 0.428970934590325, 0.418010990368202, 0.315587270306423, 0.514731952687725, 0.144373395014554, 0.593848096905276, 0.550361652392894, 0.757311524124816, 0.169781971722841, 0.417829846963286, 0.91146651096642, 0.840151912299916, 0.336678894236684, 0.0925076971761882, 0.353209213586524, 0.505065157776698, 0.979418520350009, 0.402344166766852, 0.0560892426874489, 0.332398490281776, 0.030506280483678, 0.166448161238804, 0.0309743762481958, 0.0704344878904521, 0.434927806491032, 0.674482710426673, 0.514854924753308, 0.585516086313874, 0.746886680368334, 0.521136809606105, 0.32417456433177, 0.122476802673191, 0.823901110328734, 0.754977445118129, 0.733750141458586, 0.477631961228326, 0.279780424199998, 0.330275353975594, 0.402220095507801, 0.0331362474244088, 0.938857955625281, 0.308961162110791, 0.039768025977537, 0.276454096427187, 0.0507518220692873, 0.366744362749159, 0.347430506255478, 0.538705312879756, 0.948364698095247, 0.440677372738719, 0.277127481764182, 0.261830420233309, 0.0319743093568832, 0.124868358485401, 0.488001533551142, 0.427191495662555, 0.728013194166124, 0.659969118889421, 0.337164778262377, 0.727096587652341, 0.0696419323794544, 0.861693987157196, 0.318400414194912, 0.216443409677595, 0.512548375641927, 0.357285578735173, 0.889229163294658, 0.543035649927333, 0.510798918548971, 0.590560233453289, 0.572399745695293, 0.266690234420821, 0.383943454595283, 0.99361688294448, 0.691469098441303, 0.99896499584429, 0.845753413857892, 0.886907928390428, 0.818956036353484, 0.457971467403695, 0.487093946198002, 0.76228749146685, 0.40189113211818, 0.248092987108976, 0.332653226796538, 0.229394117370248, 0.875267682597041, 0.600122341420501, 0.221825612010434, 0.190645775059238, 0.914413073798642, 0.0395654528401792, 0.754183794837445, 0.275663779815659, 0.802257802337408, 0.944307228084654, 0.257011065259576, 0.486635462380946, 0.334474149858579, 0.218721537152305, 0.800595353357494, 0.842506285058334, 0.544330153148621, 0.010595184750855, 0.0336757148616016, 0.495851775165647, 0.186647447757423, 0.555858191801235, 0.535714537603781, 0.435000111116096, 0.95906854304485, 0.0129276616498828, 0.771085542393848, 0.942945420742035, 0.646819533314556, 0.0259965704753995, 0.860606698552147, 0.784083696547896, 0.998703294899315, 0.678837856510654, 0.496942139230669, 0.416324001271278, 0.0137477868702263, 0.659451464889571, 0.963581587886438, 0.0974635290913284, 0.215850848238915, 0.729016688885167, 0.413670372450724, 0.177887552883476, 0.865083066513762, 0.831954505294561, 0.590329837985337, 0.728171345079318, 0.696531072258949, 0.351905167568475, 0.832776009337977, 0.74401311064139, 0.842459253501147, 0.575371033977717, 0.441246638773009, 0.496761611197144, 0.104383341036737, 0.307986229658127, 0.197084730258211, 0.365848952205852, 0.0381152080371976, 0.61388462362811, 0.608822415815666, 0.421242730226368, 0.075461003696546, 0.219399351160973, 0.274694567779079, 0.952492184937, 0.297489072196186, 0.801360007375479, 0.594256622483954, 0.838593422202393, 0.532689604675397, 0.484287176746875, 0.517502316040918, 0.168590852990746, 0.951992661925033, 0.490379848750308, 0.135714418021962, 0.792967187939212, 0.0378797051962465, 0.500618610763922, 0.0558760014828295, 0.0119760450907052, 0.243172378046438, 0.665593597339466, 0.894951937487349, 0.765697131399065, 0.549627005588263, 0.362301866989583, 0.700481810839847, 0.958131450228393, 0.133278971770778, 0.519576348830014, 0.257615069625899, 0.0581162446178496, 0.593921028310433, 0.497030294500291, 0.919015940511599, 0.975695187458768, 0.615137550979853, 0.466288721887395, 0.0947327807080001, 0.272690082434565, 0.44503001915291, 0.799547576345503, 0.756092163268477, 0.465563534293324, 0.00677355681546032, 0.0255012304987758, 0.218566340161487, 0.31311332876794, 0.966047712368891, 0.387194308219478, 0.508976591285318, 0.171212015440688, 0.0668710074387491, 0.551696181064472, 0.290379292098805, 0.817245903424919, 0.692718067439273, 0.0958808590658009, 0.113954037195072, 0.793890826404095, 0.142129799118266, 0.762443795800209, 0.669861795613542, 0.135129043832421), actual = c(0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.745499526849017, 0.618281947216019, 0.526423338800669, 0.575601863209158, 0.936176958261058, 0.219438234111294, 0.696261255070567, 0.863768657669425, 0.532464269781485, 0.970340725500137, 0.53836554614827, 0.176875066710636, 0.720566574018449, 0.718181093689054, 0.328918579732999, 0.605327350553125, 0.27107816352509, 0.0181633040774614, 0.762145489221439, 0.601385014131665, 0.771353341406211, 0.224830924766138, 0.94386597094126, 0.0785153112374246, 0.144368359586224, 0.713332385290414, 0.97296709055081, 0.570645660627633, 0.199405536055565, 0.822587608126923, 0.0617792683187872, 0.110537577187642, 0.821131803095341, 0.786271238932386, 0.91423281468451, 0.161845920840278, 0.39699175558053, 0.229115796508268, 0.688595491228625, 0.388115888694301, 0.840786944143474, 0.0687365783378482, 0.375664864433929, 0.570716596907005, 0.286874742479995, 0.679341854760423, 0.0726167995017022, 0.0625042738392949, 0.619155854685232, 0.310842648148537, 0.486134164500982, 0.144336099037901, 0.0581604551989585, 0.176115084905177, 0.414360082242638, 0.205268799327314, 0.934810657985508, 0.43669788679108, 0.0232717527542263, 0.430508268298581, 0.567270197439939, 0.505091103492305, 0.931037283036858, 0.781316506909207, 0.822526163421571, 0.246811743592843, 0.080093814060092, 0.841948133893311, 0.141306719509885, 0.732324595097452, 0.930567274335772, 0.86793198552914, 0.68579663685523, 0.316777187865227, 0.220331991324201, 0.939306543208659, 0.635146495187655, 0.740612066118047, 0.591072525363415, 0.687732943799347, 0.732029636856169, 0.780194329563528, 0.265543673187494, 0.0253533814102411, 0.262918512802571, 0.879435495939106, 0.0601197443902493, 0.264700108207762, 0.661616097670048, 0.612110437359661, 0.32562284078449, 0.469017032999545, 0.558957159519196, 0.589714224683121, 0.223021240439266, 0.840453540207818, 0.00321163516491652, 0.345134121365845, 0.0511110881343484, 0.180177596397698, 0.237995538162068, 0.680245544062927, 0.977238426916301, 0.911372107686475, 0.522269573993981, 0.209811202948913, 0.309487238526344, 0.586481601698324, 0.538497596280649, 0.984736623708159, 0.0236689418088645, 0.377936787670478, 0.546231482410803, 0.664931659586728, 0.426287651294842, 0.805556582985446, 0.579612896777689, 0.847713702591136, 0.654624489601701, 0.7543952295091, 0.0110853761434555, 0.317048819269985, 0.777952175121754, 0.0865423881914467, 0.15215480630286, 0.144627064466476, 0.293838104465976, 0.982907915720716, 0.0831204808782786, 0.450248656561598, 0.416339927352965, 0.948920631315559, 0.465341447852552, 0.293898950563744, 0.238634428009391, 0.217597001232207, 0.508709389017895, 0.900593637255952, 0.819648053497076, 0.455598547589034, 0.658135356381536, 0.373192951083183, 0.996592794544995, 0.569988784380257, 0.548737329430878, 0.275936890859157, 0.788456460926682, 0.884835710516199, 0.918631116161123, 0.649888909189031, 0.249872866785154, 0.50175682338886, 0.249830693937838, 0.450342760188505, 0.431751335505396, 0.969402783783153, 0.650303643429652, 0.726588089484721, 0.287370595382527, 0.999971132026985, 0.744140475988388, 0.428515642182902, 0.59870383515954, 0.278289509471506, 0.395442166831344, 0.913524742936715, 0.649712462443858, 0.419847938232124, 0.547445777803659, 0.308910594321787, 0.784663556143641, 0.109104925999418, 0.725141575792804, 0.477541008032858, 0.625916146906093, 0.427471962990239, 0.548239236464724, 0.489049347816035, 0.22861492796801, 0.661391834495589, 0.967992659192532, 0.936416121665388, 0.879790720297024, 0.712573990225792, 0.403130187653005, 0.308895935304463, 0.190283185103908, 0.522074841195717, 0.937892911490053, 0.340817776974291, 0.810777493519709, 0.633235101820901, 0.232296281028539, 0.500582947162911, 0.928898768732324, 0.622328181518242, 0.00710779684595764, 0.706540230894461, 0.563391108997166, 0.769026847090572, 0.732634214684367, 0.956593722337857, 0.97975222277455, 0.325421318411827, 0.27954148221761, 0.457026493502781, 0.59737511142157, 0.845425425563008, 0.0334713426418602, 0.338216865435243, 0.206464076414704, 0.291977043962106, 0.904797099530697, 0.755262350430712, 0.999842391116545, 0.106872346252203, 0.391352970618755, 0.524953399552032, 0.497318239882588, 0.417590231634676, 0.237308701267466, 0.996609393507242, 0.501688533229753, 0.179333982523531, 0.411300798878074, 0.0140156347770244, 0.838615789078176, 0.382408100645989, 0.874063528142869, 0.47395833209157, 0.371503892587498, 0.306506941327825, 0.764329597586766, 0.394460164243355, 0.631078758276999, 0.566746364114806, 0.987383748171851, 0.152264406671748, 0.225913664791733, 0.397586218547076, 0.951875990023836, 0.536034419434145, 0.875837278319523, 0.281390806427225, 0.460681568132713, 0.836927699623629, 0.914740334730595, 0.0621935492381454, 0.0101259278599173, 0.962863843655214, 0.95755616389215, 0.907881882041693, 0.662426240975037, 0.16911328304559, 0.161703930934891, 0.510821184143424, 0.87545733852312, 0.780997465364635, 0.412659003632143, 0.930986463092268, 0.392553582089022, 0.00522652012296021, 0.0846709501929581, 0.4808858493343, 0.637758787721395, 0.76410619285889, 0.709005798678845, 0.516312334453687, 0.069484316976741, 0.68508669687435, 0.26111781434156, 0.176079349126667, 0.121625680709258, 0.423303220421076, 0.698593605775386, 0.760476912837476, 0.312672514235601, 0.568211472127587, 0.277114322176203, 0.641921170288697, 0.00571882422082126, 0.984386943047866, 0.322766328928992, 0.885886468458921, 0.223456950392574, 0.0870994434226304, 0.225280702114105, 0.142867788672447, 0.0745835707057267, 0.882707229582593, 0.7252362431027, 0.376585858408362, 0.657243254594505, 0.208970369538292, 0.561681532068178, 0.456635301699862, 0.214112860849127, 0.368873839033768, 0.518112532794476, 0.737502391450107, 0.913004286121577, 0.209860469214618, 0.337475970387459, 0.222040509339422, 0.0926810356322676, 0.722069789655507, 0.254554850514978, 0.709698911057785, 0.0832746205851436, 0.0984752734657377, 0.945612074341625, 0.574579486157745, 0.955315110972151, 0.940411397954449, 0.0728705020155758, 0.366476827301085, 0.199661537306383, 0.899345976766199, 0.952703617047518, 0.661782949231565, 0.646178711904213, 0.487033243756741, 0.679046736098826, 0.104842915665358, 0.398541153641418, 0.803763805190101, 0.232381508918479, 0.517346795648336, 0.480467647081241, 0.668529243208468, 0.595135280396789, 0.368516652379185, 0.783243254059926, 0.625302239321172, 0.793868154753, 0.0907922340556979, 0.623699764953926, 0.373513923492283, 0.822011129232123, 0.234047871315852, 0.0351268665399402, 0.327466197777539, 0.518472274532542, 0.710718189366162, 0.358801515074447, 0.209665738744661, 0.351521053584293, 0.358646436128765, 0.664546756306663, 0.828669678419828, 0.0789457419887185, 0.725283561041579, 0.404064415721223, 0.011235038517043, 0.46803092258051, 0.254929835209623, 0.9571852106601, 0.320714532863349, 0.857278638985008, 0.828969611320645, 0.314791474491358, 0.379381025442854, 0.760832690633833, 0.0122543850447983, 0.949821440503001, 0.358060744591057, 0.243656358681619, 0.910974010592327, 0.313208322972059, 0.0768568848725408, 0.776276461081579, 0.915603202069178, 0.023158892756328, 0.121785531751812, 0.942964629502967, 0.492969694780186, 0.320930622750893, 0.254127185093239, 0.476991718402132, 0.357194340554997, 0.379751639207825, 0.0274765659123659, 0.249801566591486, 0.758616492385045, 0.544185836100951, 0.124793374212459, 0.496112318243831, 0.637672126758844, 0.250646924134344, 0.95528700388968, 0.380703018046916, 0.197924006730318, 0.797577723860741, 0.841736857546493, 0.357306177960709, 0.13767141988501, 0.370458583347499, 0.659880138933659, 0.0232839742675424, 0.545170405879617, 0.564767111092806, 0.92533233971335, 0.517717256909236, 0.0600068641360849, 0.58964483672753, 0.648821332259104, 0.862690066453069, 0.853823322337121, 0.437059257179499, 0.183645309181884, 0.484719719272107, 0.0248287082649767, 0.131641124142334, 0.185638036578894, 0.860359238926321, 0.720646530855447, 0.68250324181281, 0.960352040361613, 0.312616974813864, 0.854811139171943, 0.777061187895015, 0.997393088648096, 0.0054490405600518, 0.121333384187892, 0.638010047608986, 0.280474498169497, 0.576588652795181, 0.466776973102242, 0.148487078491598, 0.063279377296567, 0.58100353856571, 0.760398309212178, 0.884456324158236, 0.364562969654799, 0.511340600438416, 0.921033604070544, 0.843384247506037, 0.0330532311927527, 0.23892278294079, 0.955157490912825, 0.580088135786355, 0.910234122071415, 0.986846743617207, 0.953167491126806, 0.424221557099372, 0.924844544380903, 0.891870262334123, 0.407452114159241, 0.498935238691047, 0.13943237089552, 0.784796200925484, 0.847981835948303, 0.204368588048965, 0.790697892429307, 0.286922913277522, 0.0367943393066525, 0.642021450679749, 0.0282374243251979, 0.625773686217144, 0.465466843219474, 0.964882037369534, 0.309729702305049, 0.590089495293796, 0.263229244854301, 0.855955272447318, 0.4746028566733, 0.057429384207353, 0.433477527927607, 0.723609921056777, 0.0387424738146365, 0.63167194603011, 0.898258226457983, 0.112045854097232, 0.825681873830035, 0.722607828676701, 0.462176609551534, 0.153166750445962, 0.119591578375548, 0.149898730451241, 0.231899785576388, 0.526801113504916, 0.859635280910879, 0.0309439899865538, 0.160820238990709, 0.19728104560636, 0.361283675767481, 0.980787337757647, 0.59044328588061, 0.0783049631863832, 0.259970015147701, 0.699185368372127, 0.824303460307419, 0.162919738795608, 0.346625273581594, 0.353470019763336, 0.535198365105316, 0.604245314374566, 0.61014121840708, 0.453223552322015, 0.514302884927019, 0.870087351417169, 0.0747985320631415, 0.937505339737982, 0.687490778043866, 0.84208182268776, 0.583505115704611, 0.41903935582377, 0.5074552600272, 0.41782066388987, 0.855081523070112, 0.300121440086514, 0.551449909806252, 0.273157358402386, 0.415248560486361, 0.149165334645659, 0.620701763546094, 0.836005375022069, 0.206358658382669, 0.764762269100174, 0.912910816492513, 0.397522201761603, 0.134793301811442, 0.0261298383120447, 0.0393476663157344, 0.98852832778357, 0.613410020479932, 0.890108890365809, 0.828836065949872, 0.443444291129708, 0.110535922460258, 0.987074829172343, 0.371718563837931, 0.889381682500243, 0.498289703158662, 0.319847748149186, 0.534647216787562, 0.836859488626942, 0.624375216662884, 0.33412362053059, 0.29101915913634, 0.315569188678637, 0.11083711264655, 0.250982047989964, 0.349566337652504, 0.216450790641829, 0.293591002235189, 0.380815555574372, 0.651563872117549, 0.228249587118626, 0.30306156212464, 0.222226704237983, 0.649157817009836, 0.387385702924803, 0.614974782103673, 0.506039281142876, 0.934896423248574, 0.624495956115425, 0.823181748855859, 0.64176710206084, 0.392704976955429, 0.087100723059848, 0.989411575719714, 0.189646987477317, 0.959441131446511, 0.893043795600533, 0.283191426889971, 0.0280016066972166, 0.282706405268982, 0.838832108769566, 0.227170400088653, 0.196892522275448, 0.0289814143907279, 0.214056831784546, 0.0976792073342949, 0.0723276371136308, 0.51306600891985, 0.857223923085257, 0.876398746622726, 0.0171026817988604, 0.121535833226517, 0.219980447087437, 0.91376809286885, 0.0237659579142928, 0.551539634820074, 0.359360719798133, 0.849770109634846, 0.389721637358889, 0.585240041138604, 0.558792541036382, 0.282828324474394, 0.926002689870074, 0.530165658099577, 0.554562105797231, 0.625465975375846, 0.172665900317952, 0.965058342088014, 0.32040225667879, 0.648695152485743, 0.175508860731497, 0.0129101539496332, 0.769205003743991, 0.939986929064617, 0.335511581040919, 0.756645226851106, 0.920839763013646, 0.170246354304254, 0.685433549107984, 0.070778829511255, 0.881895013619214, 0.767179997637868, 0.958500291453674, 0.635619380977005, 0.885271393461153, 0.129450816428289, 0.793843833263963, 0.28785060509108, 0.281888725003228, 0.100474753417075, 0.240323015255854, 0.251939958194271, 0.938520693685859, 0.77207104745321, 0.799737654160708, 0.740762688452378, 0.980974791571498, 0.357812718953937, 0.926021592691541, 0.396241104463115, 0.110761566087604, 0.150320751592517, 0.295556941069663, 0.209637691266835, 0.699544553644955, 0.755797835532576, 0.14563049399294, 0.797069557476789, 0.804380991961807, 0.942459060577676, 0.00900021637789905, 0.974195061484352, 0.419889750890434, 0.118425358552486, 0.516090484568849, 0.923810084816068, 0.971828584792092, 0.0421802343335003, 0.980414613382891, 0.8194860892836, 0.461673930753022, 0.199974238872528, 0.82806082488969, 0.676224647788331, 0.200670629506931, 0.566188061144203, 0.122768815606833, 0.0770156513899565, 0.171075935242698, 0.393901704577729, 0.475683697499335, 0.156010087113827, 0.897203043336049, 0.0261137937195599, 0.0650358656421304, 0.61387248034589, 0.212030973983929, 0.387421184917912, 0.0588115835562348, 0.352516589453444, 0.998864780878648, 0.590562825556844, 0.651127211051062, 0.948054069885984, 0.614219923503697, 0.0216857495252043, 0.748298107646406, 0.191091479267925, 0.39837097027339, 0.763938455143943, 0.379609364084899, 0.981169711332768, 0.435610234038904, 0.390985102159902, 0.136383168632165, 0.495435139862821, 0.81067395908758, 0.0664938702248037, 0.711084390757605, 0.351219383068383, 0.833448203513399, 0.60297564859502, 0.301861795829609, 0.281822239980102, 0.90029845898971, 0.692045886535197, 0.622645230032504, 0.779389517614618, 0.937050317646936, 0.698076046770439, 0.00172811909578741, 0.712955034803599, 0.0824459351133555, 0.226571009028703, 0.464197739027441, 0.987922612810507, 0.507714507402852, 0.692533646477386, 0.977236869744956, 0.452242384199053, 0.0600767175201327, 0.82131852186285, 0.229628464207053, 0.221115070628002, 0.788887661648914, 0.284713569097221, 0.331533772638068, 0.578384603373706, 0.268756294623017, 0.795508594252169, 0.274095345521346, 0.684602448483929, 0.844870953354985, 0.454743568552658, 0.62378534860909, 0.781766569474712, 0.746995580149814, 0.259296917589381, 0.222160799661651, 0.788995784241706, 0.432083658641204, 0.653539836872369, 0.0160661910194904, 0.925937386928126, 0.181890392210335, 0.954346834914759, 0.145938199246302, 0.955842550145462, 0.78791531547904, 0.352998881600797, 0.681657612556592, 0.40930233290419, 0.68110992689617, 0.580344418995082, 0.697369687492028, 0.0604220735840499, 0.0116855585947633, 0.948115790262818, 0.917571367463097, 0.182346571935341, 0.524034207453951, 0.706614980008453, 0.763725557830185, 0.152101874118671, 0.431287073064595, 0.283670108299702, 0.9390812239144, 0.0399286367464811, 0.773525008466095, 0.912664090516046, 0.399982034461573, 0.267144101206213, 0.354939372045919, 0.984415112063289, 0.438638417283073, 0.501758319791406, 0.829888647422194, 0.744779529049993, 0.14907282195054, 0.50222363951616, 0.482050691731274, 0.916858830256388, 0.214109981898218, 0.885423683328554, 0.617584059480578, 0.981855907710269, 0.345757099566981, 0.117874414892867, 0.289151096250862, 0.745324137620628, 0.00962719367817044, 0.116698352387175, 0.768273029709235, 0.75779063696973, 0.734277363866568, 0.00787521642632782, 0.437432593433186, 0.232288087951019, 0.478873002110049, 0.579158235806972, 0.0320722092874348, 0.0136124431155622, 0.452850620960817, 0.81851626210846, 0.258829368976876, 0.465122603112832, 0.0214577121660113, 0.458574644522741, 0.593886155169457, 0.149192854762077, 0.288009112933651, 0.662069011712447, 0.691345884464681, 0.363714952254668, 0.836711849318817, 0.406635254388675, 0.688157406635582, 0.798780095530674, 0.896279249805957, 0.546065281378105, 0.879951998591423, 0.969945213058963, 0.411293211393058, 0.904180048732087, 0.955100055783987, 0.838764557614923, 0.427832141518593, 0.678925636922941, 0.0457836114801466, 0.660697101149708, 0.261202502995729, 0.134681142866611, 0.0802318423520774, 0.992108629085124, 0.970377611462027, 0.0453959230799228, 0.518910474143922, 0.800104249501601, 0.713270486332476, 0.645240848418325, 0.208959418814629, 0.574836075305939, 0.982901324285194, 0.67673290357925, 0.761334577109665, 0.624552234075963, 0.35463357786648, 0.763734474312514, 0.327574418624863, 0.649906187085435, 0.540267467964441, 0.832043343922123, 0.696405267808586, 0.363337315386161, 0.499202001374215, 0.638994951266795, 0.850396961672232, 0.0341100303921849, 0.609738322906196, 0.328693507704884, 0.810280477628112, 0.359880652977154, 0.640949701657519, 0.0781071297824383, 0.24096043407917, 0.27860268810764, 0.704762804321945, 0.458457539789379, 0.399875407805666, 0.68375059356913, 0.314044138649479, 0.311224720906466, 0.483135449467227, 0.21062646759674, 0.282410334097221, 0.287288263440132, 0.0728519395925105, 0.557308894814923, 0.11131673050113, 0.354873301228508, 0.859698485117406, 0.461816732306033, 0.290898674167693, 0.288069017929956, 0.72218172182329, 0.178758597467095, 0.410377838881686, 0.548199575394392, 0.0778949023224413, 0.738830963615328, 0.763149969046935, 0.294033785583451, 0.579663492972031, 0.973258048528805, 0.402716736774892, 0.465557143790647, 0.801910982700065, 0.892847754992545, 0.590703641064465, 0.387162252562121, 0.21962504228577, 0.393102410016581, 0.00395467691123486, 0.747129749506712, 0.0753492349758744, 0.431865508202463, 0.782705541234463, 0.164754125289619, 0.862657972145826, 0.967154066544026, 0.374882835894823, 0.662888828897849, 0.100663255900145, 0.982428345130756, 0.76938263210468, 0.922185522504151, 0.446953721577302, 0.0162124445196241, 0.544422997627407, 0.817968510091305, 0.255080016795546, 0.803949005203322, 0.837942687096074, 0.684302764479071, 0.466059405123815, 0.192452745744959, 0.172825926914811, 0.636452205944806, 0.372760166181251, 0.707360569620505, 0.593755740439519, 0.934598757186905, 0.744365781312808, 0.211241778917611, 0.135606000665575, 0.406465597217903, 0.0710720864590257, 0.561534914653748, 0.0161977501120418, 0.496966441394761, 0.373560253996402, 0.833044167142361, 0.183411916019395, 0.653433015104383, 0.904312785714865, 0.33203068212606, 0.146696390816942, 0.389431196963415, 0.0886063277721405, 0.0697898908983916, 0.819515800103545, 0.647031564731151, 0.196704372297972, 0.739600734319538, 0.253340841270983, 0.0112729610409588, 0.350792316021398, 0.595309397671372, 0.726370039861649, 0.571609137114137, 0.600160579429939, 0.245318774832413, 0.980847634607926, 0.702340181451291, 0.76302364631556, 0.35663823899813, 0.910937739303336, 0.347593426471576, 0.251286555314437, 0.618057242128998, 0.0557783266995102, 0.213472278788686, 0.106037583202124, 0.345927069429308, 0.588854792295024, 0.51489747967571, 0.462750081904233, 0.423497067531571, 0.401327223982662, 0.737822249997407, 0.484295381931588, 0.74807143304497, 0.0912053461652249, 0.959729030961171, 0.972812301246449, 0.825971130048856, 0.112712817499414, 0.0563302407972515, 0.048053833656013, 0.00196871999651194, 0.989780921023339, 0.424682869110256, 0.0218872190453112, 0.479449192760512, 0.847231727326289, 0.290990556357428, 0.241861615562811, 0.621352943126112, 0.412274414207786, 0.27726197661832, 0.818625111831352, 0.284398522926494, 0.00280296755954623, 0.41804674663581, 0.633891165489331, 0.502604109700769, 0.800843552453443, 0.762050189776346, 0.790294388541952, 0.98418531822972, 0.877309539588168, 0.672937514260411), actual = c(1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"