Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4271920        0.4141182 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4555317        0.4738230        0.4882690 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.79864368820563, 0.615910986904055, 0.741751871537417, 0.660882627591491, 0.595393938245252, 0.998012305004522, 0.781591011676937, 0.49862505029887, 0.798873293446377, 0.715429945383221, 0.904950512805954, 0.112844923278317, 0.997128018643707, 0.838484378764406, 0.212516974424943, 0.622685591224581, 0.811709427507594, 0.236657656496391, 0.123947371728718, 0.755798707948998, 0.350327150663361, 0.958272539777681, 0.875936045078561, 0.290433887625113, 0.327175233280286, 0.860721952747554, 0.613836942240596, 0.185543989529833, 0.0530895839910954, 0.748602532083169, 0.828074438264593, 0.305824952200055, 0.859393665101379, 0.0682793979067355, 0.445724526653066, 0.664580260170624, 0.940511529799551, 0.985924984095618, 0.750134325353429, 0.138079624855891, 0.984869067091495, 0.264730243710801, 0.958972907392308, 0.779898339184001, 0.586188773158938, 0.494108447572216, 0.221419078530744, 0.748657058225945, 0.938793931389228, 0.134653309127316, 0.069513468304649, 0.908011505613104, 0.498617082368582, 0.573904321994632, 0.98380480450578, 0.976347706280649, 0.124207759508863, 0.936250843806192, 0.211068191332743, 0.740260011982173, 0.35992404515855, 0.728716505691409, 0.20175660494715, 0.300700076157227, 0.999607942299917, 0.251486611086875, 0.254171518143266, 0.437672040890902, 0.695146270794794, 0.647455807542428, 0.124351236503571, 0.718111224472523, 0.246471195248887, 0.551209956873208, 0.755601565819234, 0.339503716910258, 0.30823275540024, 0.300936303799972, 0.607378213433549, 0.590304964920506, 0.0710210858378559, 0.311351096490398, 0.706498944200575, 0.60062319249846, 0.0780177055858076, 0.702392469393089, 0.0622227757703513, 0.024089916376397, 0.367842147126794, 0.0410064917523414, 0.859112511621788, 0.821666597621515, 0.151855538832024, 0.494088162668049, 0.257698852568865, 0.376790628070012, 0.360137181356549, 0.988667023135349, 0.104434709530324, 0.610683080274612, 0.187284452840686, 0.0632809032686055, 0.19799829646945, 0.968234217492864, 0.54776565823704, 0.858012937242165, 0.0273996184114367, 0.3863062732853, 0.902053974801674, 0.760204094462097, 0.632994665997103, 0.140352885471657, 0.415329163195565, 0.0485927055124193, 0.428562257671729, 0.871783988317475, 0.065427552908659, 0.957186820684001, 0.886012659873813, 0.670524129178375, 0.672139642294496, 0.303356372984126, 0.942566725192592, 0.542661159764975, 0.91475459234789, 0.670945553807542, 0.141911673592404, 0.984928494784981, 0.809955157805234, 0.159075092757121, 0.931915592867881, 0.583240311592817, 0.470124060055241, 0.302969597280025, 0.20683696353808, 0.880747735034674, 0.211230247514322, 0.453647577902302, 0.564654708141461, 0.0211682755034417, 0.155314488103613, 0.97029567300342, 0.645235208328813, 0.738037906121463, 0.28061744524166, 0.957612134050578, 0.315742092207074, 0.684442481026053, 0.0489187319763005, 0.166567364707589, 0.734759900253266, 0.949262184556574, 0.891594696091488, 0.416188274277374, 0.956154502229765, 0.473169858334586, 0.851435650140047, 0.0861963394563645, 0.775999720906839, 0.187047699233517, 0.473644926678389, 0.376678875414655, 0.0532443320844322, 0.213287513470277, 0.661280570784584, 0.950799077283591, 0.425024672644213, 0.432582357665524, 0.048844656907022, 0.306254123803228, 0.83447175566107, 0.350674120942131, 0.616406185552478, 0.113629279891029, 0.726184163242579, 0.2474448105786, 0.406856006011367, 0.934415529482067, 0.312759330496192, 0.0839071657974273, 0.246412714011967, 0.283696192316711, 0.822511055739596, 0.980777378892526, 0.485078894300386, 0.518468293128535, 0.512706130277365, 0.915531599661335, 0.179077440407127, 0.819539345102385, 0.128404252929613, 0.176901597063988, 0.79864674503915, 0.686100193299353, 0.510729084955528, 0.937230254756287, 0.0489926796872169, 0.528482012217864, 0.617716069566086, 0.824762033298612, 0.25953638763167, 0.260756572009996, 0.335197765147313, 0.87877229694277, 0.805168050574139, 0.230099610751495, 0.267055159434676, 0.980320372851565, 0.910508698550984, 0.917217157315463, 0.945083095459267, 0.902196782873943, 0.928295763209462, 0.197217318229377, 0.291416813852265, 0.298144490690902, 0.29486549529247, 0.763536726823077, 0.032311957096681, 0.53007630771026, 0.745278070215136, 0.707634606165811, 0.777427055872977, 0.829099841415882, 0.882691385224462, 0.317159292986616, 0.0721867787651718, 0.870249790605158, 0.957280671456829, 0.317696776241064, 0.902269436046481, 0.273271335521713, 0.417423816863447, 0.284802633803338, 0.800390600925311, 0.161417190684006, 0.640600958839059, 0.757229282520711, 0.299336273223162, 0.15812994283624, 0.905725048622116, 0.751602372387424, 0.570974895032123, 0.83140837890096, 0.197386514395475, 0.723590921144933, 0.668708590092137, 0.306353005114943, 0.634456377476454, 0.603971196571365, 0.0777932233177125, 0.00331660150550306, 0.0906740827485919, 0.559918138664216, 0.680079138837755, 0.0761662910226732, 0.238048364873976, 0.263149586506188, 0.168850685236976, 0.570233406964689, 0.714891793206334, 0.631854656385258, 0.136942790122703, 0.343448484316468, 0.677741973195225, 0.6045031840913, 0.623343745945022, 0.0329397139139473, 0.380446744849905, 0.990770555101335, 0.882456937571988, 0.254578380845487, 0.712348079076037, 0.840291424887255, 0.70771186822094, 0.907224831636995, 0.632087350822985, 0.0876549491658807, 0.407071427442133, 0.0230108411051333, 0.690584708703682, 0.818263752618805, 0.218126423889771, 0.196025728015229, 0.233000146690756, 0.829688199562952, 0.36671277997084, 0.719403184019029, 0.988815161399543, 0.907379053765908, 0.00599581841379404, 0.525982118910179, 0.221735494676977, 0.964812430785969, 0.722260343842208, 0.0380577228497714, 0.124577088747174, 0.281581219518557, 0.566186214797199, 0.763329857261851, 0.904849571641535, 0.216451924527064, 0.761165135307238, 0.0391104584559798, 0.545775862643495, 0.176605488639325, 0.775025910465047, 0.473269278183579, 0.209771629888564, 0.77405680459924, 0.0828774510882795, 0.0388801842927933, 0.648653275100514, 0.572201198432595, 0.765624180668965, 0.470312841236591, 0.357303636381403, 0.239023790927604, 0.562857547774911, 0.681373572908342, 0.758440656354651, 0.100107122212648, 0.0497851681429893, 0.215152053628117, 0.565403718501329, 0.777648263610899, 0.00584296998567879, 0.715950350044295, 0.902140786172822, 0.992553657386452, 0.560535660246387, 0.127032730961218, 0.507125907344744, 0.524460792308673, 0.198080424685031, 0.624006943078712, 0.117909203516319, 0.345131530426443, 0.650160167599097, 0.033661384601146, 0.839514363091439, 0.401957821799442, 0.805527353659272, 0.714905763510615, 0.637326088966802, 0.892589724157006, 0.312230094335973, 0.357233062386513, 0.167302697664127, 0.650631760247052, 0.799647520761937, 0.205552795203403, 0.209775569150224, 0.700377976987511, 0.564372785855085, 0.951510842656717, 0.323699287604541, 0.357507957378402, 0.161531487479806, 0.886040369281545, 0.120058291358873, 0.593775260262191, 0.0903037704993039, 0.889043096220121, 0.534756656968966, 0.890248960349709, 0.4760697116144, 0.893179018050432, 0.0687102507799864, 0.0565821637865156, 0.829325751401484, 0.0202233018353581, 0.297433387488127, 0.550448543624952, 0.713967914460227, 0.94863646873273, 0.165478500537574, 0.227397356880829, 0.209388276329264, 0.543611766770482, 0.934549423865974, 0.787653767969459, 0.703320413827896, 0.639158786972985, 0.993963520275429, 0.33450149348937, 0.717722497414798, 0.116671368479729, 0.946603951742873, 0.303911518072709, 0.316095932386816, 0.732315445551649, 0.114385632332414, 0.977241774089634, 0.524655687157065, 0.147764598950744, 0.15704985964112, 0.849257367895916, 0.496902671875432, 0.021449952153489, 0.569627375109121, 0.62160624563694, 0.644502384122461, 0.973710920196027, 0.796240251278505, 0.585536693921313, 0.378186248941347, 0.360203797696158, 0.262806277954951, 0.960501320660114, 0.939046774990857, 0.80805940926075, 0.0101159864570946, 0.829729883000255, 0.590517163043842, 0.637757950462401, 0.554006943944842, 0.785380583489314, 0.582505685277283, 0.758203235222027, 0.884167271666229, 0.58816138538532, 0.73358664335683, 0.157663027988747, 0.798402608837932, 0.995494665345177, 0.0538167082704604, 0.218172704102471, 0.811509562656283, 0.942649400327355, 0.662048186175525, 0.782603829167783, 0.776838878402486, 0.239810143597424, 0.417340486776084, 0.89802160509862, 0.621800035936758, 0.266423173481598, 0.105660188244656, 0.503172872588038, 0.791724156122655, 0.0546350656077266, 0.415156408911571, 0.242582413135096, 0.432793715735897, 0.968208857811987, 0.29509644000791, 0.225050361594185, 0.704300643876195, 0.509473910555243, 0.287106909556314, 0.81969459191896, 0.219794602366164, 0.647130586439744, 0.225099841598421, 0.75611151731573, 0.497801680117846, 0.94136628927663, 0.590351708699018, 0.0835532634519041, 0.448786376276985, 0.504761802265421, 0.394682933576405, 0.0319264519494027, 0.876486170105636, 0.0214265086688101, 0.724141861079261, 0.249023373937234, 0.570100128650665, 0.19614724535495, 0.663985653780401, 0.137901088921353, 0.679461414925754, 0.73206175211817, 0.925114904530346, 0.583649225533009, 0.52456294815056, 0.239412564318627, 0.355250503402203, 0.160454920493066, 0.686846404103562, 0.093091124901548, 0.164463967783377, 0.279425800777972, 0.0348363784141839, 0.250065098283812, 0.249625831842422, 0.868917553452775, 0.752011899603531, 0.357555381488055, 0.190080428030342, 0.577936520799994, 0.670795845100656, 0.440213041845709, 0.0283163897693157, 0.617808168521151, 0.352797070052475, 0.464673031121492, 0.440725078107789, 0.739647322334349, 0.938918504398316, 0.540133170317858, 0.157365513266996, 0.527411749353632, 0.285390785196796, 0.294752289541066, 0.755562158999965, 0.785379340406507, 0.689614093396813, 0.790777289308608, 0.727558883139864, 0.752707826904953, 0.0819787152577192, 0.509436497231945, 0.171560463029891, 0.62540783546865, 0.245226954109967, 0.725420006550848, 0.277645809343085, 0.642081251833588, 0.841101272031665, 0.82535591092892, 0.751517300959677, 0.457344924099743, 0.0633788439445198, 0.634850870585069, 0.804634707979858, 0.763804044341668, 0.417577635496855, 0.693609653506428, 0.818159315967932, 0.951251382706687, 0.885002905968577, 0.87966965441592, 0.512144374428317, 0.722735658520833, 0.73413198068738, 0.619397712405771, 0.375020558945835, 0.895139712141827, 0.0161252804100513, 0.34297350095585, 0.345150280278176, 0.842539642238989, 0.716129328822717, 0.775844779564068, 0.338053956162184, 0.174478025408462, 0.227571312803775, 0.30444646673277, 0.171827153069898, 0.618770770495757, 0.842118998989463, 0.619470843346789, 0.611920282477513, 0.133989564841613, 0.428291014162824, 0.762979445280507, 0.185993060702458, 0.15278510376811, 0.00643043988384306, 0.478772966656834, 0.43885632394813, 0.462182472459972, 0.0115102175623178, 0.0989790805615485, 0.95747495489195, 0.51690874574706, 0.175885879900306, 0.1783885685727, 0.708376009017229, 0.359741390449926, 0.626550038345158, 0.243645736714825, 0.426164501113817, 0.0604578494094312, 0.813802510732785, 0.014913191087544, 0.0723284941632301, 0.44491639919579, 0.547855462646112, 0.444677086081356, 0.977900095516816, 0.358557128114626, 0.341477423207834, 0.347296668682247, 0.84424779494293, 0.0967623279429972, 0.9716338342987, 0.218495706096292, 0.668147104326636, 0.666889010230079, 0.025107643334195, 0.235692237736657, 0.129925780231133, 0.0485168087761849, 0.39254348538816, 0.464620269136503, 0.0928077271673828, 0.696243053069338, 0.959347690921277, 0.294904073700309, 0.168585570529103, 0.907044860301539, 0.755002956604585, 0.423633241327479, 0.891009227139875, 0.484510005218908, 0.954348838422447, 0.505833949660882, 0.973290908150375, 0.917412639595568, 0.124272913439199, 0.99299838882871, 0.040836168685928, 0.758157436270267, 0.124565099366009, 0.701981849269941, 0.471806765301153, 0.14713467285037, 0.407876786077395, 0.895832732785493, 0.367556332610548, 0.903767415089533, 0.572491918690503, 0.625481794588268, 0.52414429327473, 0.871792713878676, 0.533358053304255, 0.973424232099205, 0.351979215163738, 0.382907326566055, 0.801348184933886, 0.0480383122339845, 0.949784988770261, 0.305233035935089, 0.705410548485816, 0.316746315453202, 0.108940004836768, 0.0157101135700941, 0.761235881131142, 0.356408858206123, 0.656765627907589, 0.859743119915947, 0.749682535184547, 0.492501283995807, 0.00774420821107924, 0.394004796165973, 0.184360339771956, 0.671751874731854, 0.0617172801867127, 0.368997217388824, 0.248979259980842, 0.597943669883534, 0.643533939495683, 0.110892777564004, 0.937563134822994, 0.588460284285247, 0.205865249503404, 0.0858919736929238, 0.215709282550961, 0.368579005822539, 0.266013889806345, 0.81249448354356, 0.789600322023034, 0.414095397107303, 0.680066913831979, 0.816395352827385, 0.458309825975448, 0.126134885940701, 0.0814956466201693, 0.804334650980309, 0.633811046835035, 0.361793918302283, 0.573208919726312, 0.659834752324969, 0.992486398667097, 0.202989707002416, 0.341927824309096, 0.585299710743129, 0.255385010968894, 0.256318834843114, 0.196123594650999, 0.0217319768853486, 0.741698523750529, 0.483743336051702, 0.4666486766655, 0.940335794584826, 0.286542282439768, 0.358845410868526, 0.813004471128806, 0.446431763470173, 0.494813849916682, 0.0598472550045699, 0.167906378395855, 0.950975015293807, 0.242373366141692, 0.812048375606537, 0.241798355476931, 0.96606079232879, 0.826084400992841, 0.0170215007383376, 0.886608577566221, 0.727127401158214, 0.998979227850214, 0.342943059513345, 0.0408753140363842, 0.951215128181502, 0.831882345955819, 0.715867405757308, 0.924541524611413, 0.592310620471835, 0.247927192598581, 0.477553742704913, 0.959846657933667, 0.565747204236686, 0.721908564213663, 0.149263312574476, 0.351261299336329, 0.834361695451662, 0.752684771548957, 0.400205173296854, 0.860373213421553, 0.488385146483779, 0.651151880389079, 0.233343616127968, 0.981289493152872, 0.572809238918126, 0.711048158351332, 0.542376924306154, 0.174064637394622, 0.154548248741776, 0.788807484786958, 0.680201205890626, 0.717335738008842, 0.898457868024707, 0.450950071681291, 0.778156498912722, 0.594185300404206, 0.0596352657303214, 0.837723539210856, 0.0954613322392106, 0.496951414039358, 0.936058473773301, 0.939814500976354, 0.365381359355524, 0.818908290704712, 0.02534039481543, 0.0478696306236088, 0.0700803303625435, 0.200329261366278, 0.0334819625131786, 0.619238577783108, 0.653360972180963, 0.155475873267278, 0.144417630042881, 0.420250667724758, 0.878358770394698, 0.0201632417738438, 0.922516017919406, 0.384488740004599, 0.547323620645329, 0.023481797426939, 0.646452745655552, 0.821848378749564, 0.316547181457281, 0.341919905971736, 0.125023714499548, 0.270245483145118, 0.686056585749611, 0.850195951294154, 0.440936895553023, 0.626102376496419, 0.152860066154972, 0.272932194871828, 0.512120033614337, 0.434382783714682, 0.972882936010137, 0.479357270989567, 0.870813105255365, 0.0876359411049634, 0.748438798822463, 0.986221893923357, 0.77701893215999, 0.0205119457095861, 0.589361367514357, 0.620172799332067, 0.167457216186449, 0.429325563600287, 0.229960005963221, 0.94034897419624, 0.0457726137246937, 0.1211104656104, 0.734415316022933, 0.203391101676971, 0.690347353927791, 0.010480736149475, 0.140770923346281, 0.439186749979854, 0.995242131641135, 0.792316687526181, 0.512060526292771, 0.806837644195184, 0.747541547287256, 0.990606215782464, 0.411310538416728, 0.164739098167047, 0.059611548203975, 0.864044319139794, 0.354227202013135, 0.611529621528462, 0.117778103565797, 0.642209644895047, 0.656043858034536, 0.866752204485238, 0.0132214722689241, 0.623314848169684, 0.778135108761489, 0.622050619218498, 0.841379725141451, 0.241573064122349, 0.853473668219522, 0.323357409099117, 0.102826437214389, 0.605829067295417, 0.824693570146337, 0.728465036023408, 0.474521290510893, 0.153989949962124, 0.300446776207536, 0.597064240835607, 0.58985984721221, 0.683754251338542, 0.809049943462014, 0.995479005388916, 0.63894664356485, 0.768334931693971, 0.772133355261758, 0.612328664166853, 0.362709248904139, 0.983786277938634, 0.201630045892671, 0.390800280496478, 0.535555624868721, 0.14601603616029, 0.245423063402995, 0.139824402984232, 0.561915197875351, 0.57477435702458, 0.404924469068646, 0.767815715633333, 0.793878873577341, 0.998118286952376, 0.805652173236012, 0.507264060201123, 0.866843626368791, 0.538766882847995, 0.638693561544642, 0.741393161006272, 0.566759552806616, 0.200428755953908, 0.960761794820428, 0.993584517622367, 0.776376508176327, 0.255777604877949, 0.79923167778179, 0.118224874138832, 0.828974452568218, 0.701803785515949, 0.744230775162578, 0.750775919528678, 0.389180339174345, 0.60982058220543, 0.613403771538287, 0.692082067718729, 0.586599143221974, 0.0669759132433683, 0.740907471859828, 0.0980296789202839, 0.26802286086604, 0.458795211045071, 0.457816934445873, 0.307443518890068, 0.590817395597696, 0.0988885646220297, 0.417325070127845, 0.696979366010055, 0.955637414008379, 0.140495619969442, 0.0412351465784013, 0.766646774485707, 0.866853113519028, 0.560252058552578, 0.323786888737231, 0.954076991882175, 0.415647700196132, 0.724863734329119, 0.291215395787731, 0.54282883531414, 0.00846438040025532, 0.293570986948907, 0.107626226730645, 0.658003009157255, 0.313620708649978, 0.445322775980458, 0.56365314964205, 0.911821855697781, 0.886113927001134, 0.945046081440523, 0.692500427830964, 0.921257078880444, 0.208378992276266, 0.771948522655293, 0.780445853481069, 0.430046293186024, 0.977674719877541, 0.276701369090006, 0.32569448556751, 0.119883965933695, 0.888728839345276, 0.51895520137623, 0.781994359334931, 0.178628076566383, 0.863597315037623, 0.541068040765822, 0.37583724479191, 0.991383923683316, 0.683973834849894, 0.453414329793304, 0.0425859452225268, 0.134136675857008, 0.694273056229576, 0.545199677348137, 0.672196653671563, 0.488163583446294, 0.196524064056575, 0.253902149619535, 0.373476076638326, 0.144997763214633, 0.477275999030098, 0.72596106887795, 0.252438940806314, 0.773968782741576, 0.600667474325746, 0.222541405120865, 0.642109425505623, 0.062164822826162, 0.856323117157444, 0.965855985647067, 0.233029679162428, 0.054861958604306, 0.319903781404719, 0.509439690504223, 0.803903623716906, 0.222037855768576, 0.0258205193094909, 0.381659967126325, 0.774381130700931, 0.927933039143682, 0.39873779611662, 0.566023091319948, 0.617794569814578, 0.909445670899004, 0.084279790520668, 0.761900017270818, 0.942192886257544, 0.840758621459827, 0.320135575486347, 0.251802457496524, 0.666317639639601, 0.795004871441051, 0.123501555295661, 0.670846395660192, 0.281721748411655, 0.71626135869883, 0.330787577200681, 0.482345407828689, 0.171302821487188, 0.539134410908446, 0.467904643854126, 0.890800647204742, 0.558602558914572, 0.781803771620616, 0.65140892774798, 0.669348276918754, 0.259314436232671, 0.153500309679657, 0.412148233968765, 0.189597473945469, 0.862395555246621, 0.0338186211884022, 0.381579775363207, 0.150860578753054, 0.87625563214533, 0.458980438299477, 0.257856774842367, 0.87117165257223, 0.909040852915496, 0.954850018955767, 0.691566307563335), actual = c(0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.993740220321342, 0.11688754404895, 0.794659111183137, 0.0340200758073479, 0.225672696018592, 0.509318348718807, 0.413772301981226, 0.201548545155674, 0.0832191887311637, 0.826613414566964, 0.108008621027693, 0.673899069894105, 0.906098794890568, 0.00184341496787965, 0.532532354118302, 0.64844246651046, 0.910044211428612, 0.303919296013191, 0.711487154243514, 0.880167688941583, 0.87556149601005, 0.359740794403479, 0.476650227094069, 0.401025978382677, 0.727979711955413, 0.232752727111802, 0.215042361058295, 0.304609905695543, 0.173500159755349, 0.743995974771678, 0.142467156983912, 0.892729887738824, 0.999765408458188, 0.37448011431843, 0.666803161147982, 0.132168161682785, 0.831661796895787, 0.881425252184272, 0.403194159036502, 0.647491902345791, 0.243223180994391, 0.862923220265657, 0.117039168952033, 0.500195036176592, 0.219840993639082, 0.99121973128058, 0.240547517314553, 0.0171740362420678, 0.419150015106425, 0.795123498188332, 0.497996849007905, 0.323937913170084, 0.237941147526726, 0.851556512061507, 0.00184329110197723, 0.477156220003963, 0.803489002631977, 0.514650419121608, 0.93786209425889, 0.930179865565151, 0.337428647326306, 0.622420770814642, 0.338918514316902, 0.723998950328678, 0.705652368022129, 0.737172549124807, 0.467662349576131, 0.167372063500807, 0.258555671432987, 0.23690521507524, 0.437764535890892, 0.456393052823842, 0.71206004708074, 0.785464063985273, 0.506142207188532, 0.516603428171948, 0.582917458377779, 0.705443114740774, 0.611266297986731, 0.41479765670374, 0.997242054436356, 0.265957299619913, 0.689639273099601, 0.921777576906607, 0.0200506045948714, 0.343057972146198, 0.948100393870845, 0.33545723115094, 0.479987758677453, 0.467959709232673, 0.222870093770325, 0.207177818752825, 0.451722478261217, 0.778135374421254, 0.601837323280051, 0.164921385934576, 0.308940488379449, 0.562343525001779, 0.836848278064281, 0.510561016388237, 0.064461946953088, 0.190447228960693, 0.536468772450462, 0.643295392394066, 0.778758117929101, 0.430045455927029, 0.45243999757804, 0.899890104075894, 0.357003234094009, 0.549228687072173, 0.0300202530343086, 0.608713017776608, 0.157640767749399, 0.643864295911044, 0.707241857657209, 0.260856463806704, 0.485863081878051, 0.221970620565116, 0.339194079861045, 0.944684826070443, 0.452916440553963, 0.997638326371089, 0.71222853125073, 0.635392463300377, 0.344455849612132, 0.00151976547203958, 0.557327206712216, 0.0756828703451902, 0.0555577545892447, 0.53035295708105, 0.996030455688015, 0.713173742871732, 0.733955907635391, 0.841952210990712, 0.796796691603959, 0.184617478167638, 0.168903116835281, 0.412842389196157, 0.990481727523729, 0.440233387285843, 0.286807935684919, 0.179017888847739, 0.16098947217688, 0.418689631624147, 0.482434543315321, 0.0525596386287361, 0.849612177116796, 0.112075552111492, 0.482827982865274, 0.242019617697224, 0.798586996505037, 0.44717115117237, 0.689753505866975, 0.832335822284222, 0.841758212540299, 0.298492463771254, 0.481222397880629, 0.112927054986358, 0.790555072017014, 0.649507272057235, 0.829689529724419, 0.131183146499097, 0.686338284285739, 0.911967183463275, 0.652963928179815, 0.197354226838797, 0.788097505923361, 0.432250014971942, 0.238685391843319, 0.135401244740933, 0.535671902121976, 0.70890648313798, 0.793850105721503, 0.481955118710175, 0.263433452462777, 0.808121865382418, 0.658945688046515, 0.879719186574221, 0.163457221351564, 0.129131950438023, 0.359526236075908, 0.331757599953562, 0.00204643327742815, 0.0570002652239054, 0.609076199820265, 0.370497415773571, 0.791982101276517, 0.366485038772225, 0.681304355617613, 0.554154558572918, 0.342466448666528, 0.913554138736799, 0.117261261679232, 0.344815730815753, 0.235570958349854, 0.916336174122989, 0.824352938681841, 0.017395929666236, 0.517489248886704, 0.191721272654831, 0.71928130579181, 0.712960058823228, 0.124612426385283, 0.960401902440935, 0.0896487683057785, 0.700601890450343, 0.346605717437342, 0.878375040600076, 0.453943686094135, 0.553241927409545, 0.0439603466074914, 0.746903127757832, 0.00881988788023591, 0.653133860556409, 0.529013301245868, 0.348618597025052, 0.107065187534317, 0.422785000642762, 0.715008921455592, 0.167184324469417, 0.224002818576992, 0.0784152334090322, 0.154575749998912, 0.943774230545387, 0.79258764279075, 0.393843403551728, 0.155162350041792, 0.816362433135509, 0.418329448904842, 0.14077237341553, 0.432012428762391, 0.61726038181223, 0.614291639532894, 0.167696425225586, 0.47904509329237, 0.646534638246521, 0.178998401388526, 0.852362590376288, 0.489674455486238, 0.56487406953238, 0.635519263567403, 0.165523613803089, 0.786842865869403, 0.529945325106382, 0.211646181065589, 0.660502027953044, 0.884236061014235, 0.248464825330302, 0.340221379417926, 0.857718026265502, 0.189895294839516, 0.0564869472291321, 0.0635608304291964, 0.385178437456489, 0.0439865840598941, 0.361691885860637, 0.984624888747931, 0.960787317249924, 0.428357990924269, 0.72699155821465, 0.297977658454329, 0.661431762389839, 0.640085734426975, 0.538550221594051, 0.228115010773763, 0.878049101447687, 0.608993545873091, 0.575318133691326, 0.889974297489971, 0.967413487378508, 0.739969909656793, 0.664185672532767, 0.604115828406066, 0.811334358528256, 0.150466379709542, 0.893002899596468, 0.842750805895776, 0.210648188833147, 0.250124182552099, 0.669484204845503, 0.11958915181458, 0.51287800911814, 0.932126434054226, 0.190535892732441, 0.747884969925508, 0.310181580018252, 0.574899723986164, 0.0848298720084131, 0.11307004163973, 0.00914536882191896, 0.0264506277162582, 0.442307441495359, 0.802950922865421, 0.392112503992394, 0.928521171445027, 0.400118201505393, 0.0222935758065432, 0.636895483359694, 0.695248676231131, 0.743274761131033, 0.797922045690939, 0.107201674720272, 0.880743487970904, 0.497369969030842, 0.283480263082311, 0.763723470037803, 0.710399999981746, 0.29550692695193, 0.286625965731218, 0.358799246372655, 0.816546568647027, 0.910611364757642, 0.316630437737331, 0.659111390123144, 0.897948039229959, 0.227994595887139, 0.430123605998233, 0.31394503870979, 0.188327756244689, 0.16554482281208, 0.258234188193455, 0.930199791211635, 0.929306278005242, 0.2243993782904, 0.1976246482227, 0.944976193830371, 0.410266563296318, 0.551674598129466, 0.903573065763339, 0.334257317939773, 0.520708135096356, 0.220027305185795, 0.232545564649627, 0.198898417875171, 0.317125633126125, 0.827709729317576, 0.0762827056460083, 0.182890887372196, 0.885948052164167, 0.844898263458163, 0.606040593003854, 0.551678686169907, 0.234544906998053, 0.0839221037458628, 0.812131576007232, 0.339759743539616, 0.579180265078321, 0.620539366267622, 0.767249128781259, 0.671167158754542, 0.0919129364192486, 0.686386035056785, 0.947475222405046, 0.886685240548104, 0.316248222254217, 0.960944870021194, 0.349449390545487, 0.263312943279743, 0.808236639015377, 0.48349298001267, 0.916328457882628, 0.778255172539502, 0.173932161647826, 0.969756601843983, 0.431665432406589, 0.0592608002480119, 0.126874121604487, 0.087685109814629, 0.490675595588982, 0.935490182833746, 0.480139845749363, 0.0101318384986371, 0.353119862265885, 0.564962859032676, 0.537367449840531, 0.0618502821307629, 0.177642905386165, 0.452245991211385, 0.355147309834138, 0.0820970663335174, 0.376095658866689, 0.800731641240418, 0.767180768074468, 0.139573927037418, 0.773729938082397, 0.273679915815592, 0.216852721758187, 0.467672170838341, 0.75382993533276, 0.734476524172351, 0.187814078759402, 0.0891159097664058, 0.999415366211906, 0.839497424196452, 0.491938798455521, 0.393556629307568, 0.633037826279178, 0.0778727806173265, 0.729592012241483, 0.456701978808269, 0.0580119006335735, 0.926611081231385, 0.819894392276183, 0.69930643774569, 0.967675460269675, 0.542912555392832, 0.22642050194554, 0.727305173873901, 0.845586505485699, 0.248444013064727, 0.0639912467449903, 0.754341849125922, 0.350660516880453, 0.628920911345631, 0.935166480019689, 0.50907637970522, 0.363604301586747, 0.12558013596572, 0.747804838931188, 0.968491085572168, 0.638405920471996, 0.420262037776411, 0.291742473142222, 0.956607627216727, 0.0720092761330307, 0.8221741898451, 0.960073072696105, 0.478866000659764, 0.279220183379948, 0.805667293258011, 0.757905309088528, 0.703817848581821, 0.847775970818475, 0.266132401768118, 0.805081748403609, 0.862124786246568, 0.903930358355865, 0.91649861773476, 0.462526507908478, 0.917756441049278, 0.641521190525964, 0.895636388566345, 0.898066678317264, 0.194889635313302, 0.220339125487953, 0.858602668391541, 0.560575758572668, 0.519132774323225, 0.89664422837086, 0.540520496899262, 0.392299128230661, 0.459906356176361, 0.527945747599006, 0.499364017741755, 0.479823577450588, 0.563602740876377, 0.754661511396989, 0.730527534615248, 0.955636978615075, 0.736746055306867, 0.593667703680694, 0.920631162123755, 0.640848143259063, 0.246359821641818, 0.142348378896713, 0.987819991307333, 0.799404467688873, 0.0404760574456304, 0.859861079137772, 0.638517334125936, 0.553961785743013, 0.52143298019655, 0.203302876558155, 0.566171281738207, 0.0124812559224665, 0.755152417812496, 0.725317352451384, 0.177903307136148, 0.413745973724872, 0.216416665818542, 0.281441782135516, 0.26578696235083, 0.578739789081737, 0.42305976874195, 0.441734214313328, 0.331964645301923, 0.961104719666764, 0.328254837542772, 0.290091688279063, 0.26442523021251, 0.443382252706215, 0.99533614050597, 0.120593891479075, 0.366009873570874, 0.0654499777592719, 0.367643288103864, 0.133828686317429, 0.47334135347046, 0.612535126507282, 0.359218036523089, 0.385474426904693, 0.978948323987424, 0.591451514977962, 0.444918179884553, 0.875350571237504, 0.545135592808947, 0.189308981411159, 0.314359308453277, 0.878159898798913, 0.00488844746723771, 0.399952075444162, 0.717598067596555, 0.957123788772151, 0.650401799939573, 0.113053298089653, 0.947140641044825, 0.762603918090463, 0.474720275960863, 0.403040785109624, 0.794586283387616, 0.859208945184946, 0.973768711555749, 0.230952866375446, 0.945071818307042, 0.761619028635323, 0.836256941780448, 0.770037325331941, 0.315111600328237, 0.0017998400144279, 0.890958073781803, 0.722487493418157, 0.287860334850848, 0.336464753141627, 0.405663948040456, 0.556528734974563, 0.88370790774934, 0.817473731935024, 0.178582621039823, 0.800319423433393, 0.0702350847423077, 0.374775454169139, 0.841162208933383, 0.721598583040759, 0.224253282183781, 0.479963967809454, 0.413031205534935, 0.906310790451244, 0.0408493741415441, 0.907558331731707, 0.653320746263489, 0.61576083721593, 0.984869042644277, 0.882699384586886, 0.493582271737978, 0.691420130198821, 0.378671637736261, 0.170410436345264, 0.0755852316506207, 0.341807311400771, 0.23860820941627, 0.611737187486142, 0.0763571066781878, 0.336145190754905, 0.903850648086518, 0.675929788965732, 0.330092069227248, 0.74271415756084, 0.148904734291136, 0.104845196241513, 0.443087441846728, 0.056491173338145, 0.0563646683003753, 0.652053868630901, 0.173939345171675, 0.99654317740351, 0.272942833602428, 0.914024142082781, 0.748357894131914, 0.915671509224921, 0.271468492690474, 0.170774724334478, 0.471864772727713, 0.265259488252923, 0.420184797840193, 0.410631848499179, 0.447430561995134, 0.349315705010667, 0.375271644443274, 0.602613929426298, 0.356044145533815, 0.680511026177555, 0.51430396665819, 0.771171576576307, 0.146542696515098, 0.542080166516826, 0.498500349465758, 0.48046876792796, 0.331144715892151, 0.295921101234853, 0.54121075454168, 0.738227810943499, 0.419936740072444, 0.99167166184634, 0.76617862842977, 0.398121580481529, 0.581559038721025, 0.487009619129822, 0.0592870495747775, 0.586661936482415, 0.705867795506492, 0.335720211733133, 0.199801292270422, 0.535160906612873, 0.60190770495683, 0.855207501910627, 0.655233699828386, 0.477066739695147, 0.119017250835896, 0.850213410332799, 0.28990067797713, 0.378282773774117, 0.0611509236041456, 0.758176988223568, 0.331690001301467, 0.0280979399103671, 0.971993783023208, 0.30475552729331, 0.457571520935744, 0.657876560697332, 0.456929759588093, 0.176810815464705, 0.143188444664702, 0.062873027054593, 0.300671810051426, 0.257822781102732, 0.331780193373561, 0.164363925345242, 0.266627062112093, 0.914024054072797, 0.558766781818122, 0.756637705722824, 0.768200942082331, 0.25402461970225, 0.490147373639047, 0.157244012225419, 0.395852174144238, 0.838453776901588, 0.943282225867733, 0.72345964726992, 0.843102297047153, 0.793593322858214, 0.488702081842348, 0.772573295515031, 0.414072440005839, 0.516475301235914, 0.0722801419906318, 0.575548522407189, 0.866070299176499, 0.0853564112912863, 0.332549572922289, 0.212071627611294, 0.104519591201097, 0.796802232274786, 0.302276444388554, 0.315518076531589, 0.507231596624479, 0.718236748827621, 0.458857755176723, 0.108125247759745, 0.596453465288505, 0.919131155824289, 0.316156773129478, 0.438520143041387, 0.516720597166568, 0.617550742579624, 0.973970071645454, 0.266223252983764, 0.95265701180324, 0.396926551591605, 0.311763424891979, 0.678846423514187, 0.650660417973995, 0.904738565674052, 0.425881804898381, 0.264759180834517, 0.285982307279482, 0.794849845813587, 0.579373233485967, 0.414974891813472, 0.400691272923723, 0.614783158292994, 0.575891222106293, 0.921693471958861, 0.900989694753662, 0.433596559101716, 0.281984871020541, 0.246629946632311, 0.675329068209976, 0.429144185502082, 0.860761551884934, 0.952844544779509, 0.765804210910574, 0.082166439620778, 0.641573461936787, 0.875781676499173, 0.0898165702819824, 0.709573471918702, 0.547904985956848, 0.972944014240056, 0.351140073034912, 0.713024103082716, 0.593876594677567, 0.815330244600773, 0.683432412799448, 0.261312322691083, 0.626003510318696, 0.655053090769798, 0.275627447525039, 0.649564256891608, 0.617334201233461, 0.907829067204148, 0.737764226738364, 0.755553076509386, 0.415691993199289, 0.692556144203991, 0.720624704379588, 0.13622958236374, 0.944623921066523, 0.185837593860924, 0.828517436049879, 0.43859682045877, 0.285121047170833, 0.00612986879423261, 0.868323055095971, 0.610276014078408, 0.552258652867749, 0.443024394335225, 0.142908719368279, 0.32000320032239, 0.251798716373742, 0.54857554170303, 0.0260150865651667, 0.123872081981972, 0.977800903376192, 0.900912115350366, 0.0652637998573482, 0.178459127666429, 0.903293117647991, 0.219142698682845, 0.751504606800154, 0.930424910970032, 0.255330795887858, 0.527358793420717, 0.234779036836699, 0.28922089538537, 0.502848276635632, 0.0542208265978843, 0.958441762719303, 0.355464478954673, 0.433944269316271, 0.0963622306007892, 0.60663594561629, 0.993650564225391, 0.455350078642368, 0.514395677717403, 0.805013557663187, 0.00894968234933913, 0.276808144524693, 0.881888340460137, 0.988781592808664, 0.253800892503932, 0.540408081375062, 0.75655996799469, 0.837911875452846, 0.0153539685998112, 0.974056628765538, 0.194690643111244, 0.358318589394912, 0.236957386834547, 0.90530275227502, 0.359298442257568, 0.684348554583266, 0.887917719082907, 0.307494120905176, 0.266274860827252, 0.177348109893501, 0.367585413856432, 0.431546277599409, 0.00237773498520255, 0.281298577785492, 0.223814291413873, 0.786465947981924, 0.36085081868805, 0.0413383864797652, 0.807970786467195, 0.415876125684008, 0.755816655466333, 0.687858884688467, 0.612968580797315, 0.951497421367094, 0.920731990598142, 0.00651912298053503, 0.861340226139873, 0.681035279762, 0.276746733812615, 0.0262023468967527, 0.561830018414184, 0.630292999790981, 0.966576418839395, 0.804343611234799, 0.697212727740407, 0.22167146182619, 0.182903578039259, 0.850646232254803, 0.564471611753106, 0.238164703128859, 0.536225819494575, 0.514291091123596, 0.32009445480071, 0.529808753635734, 0.987304157111794, 0.599982197862118, 0.0123529357369989, 0.488555799704045, 0.0825197531376034, 0.398801174014807, 0.632233547279611, 0.769571464508772, 0.119872605893761, 0.27892420697026, 0.594721384579316, 0.111506968038157, 0.20950966165401, 0.161603650543839, 0.239701367216185, 0.214633728610352, 0.721879105316475, 0.141482765553519, 0.358872435986996, 0.0473110601305962, 0.424580475548282, 0.780955184716731, 0.720635664183646, 0.015402895398438, 0.403100124793127, 0.45558060053736, 0.470296420855448, 0.0356070348061621, 0.172117135487497, 0.667158646509051, 0.798550489125773, 0.0634882766753435, 0.664411849342287, 0.615532583091408, 0.543357316404581, 0.914897223236039, 0.28175389720127, 0.328242104966193, 0.821053462103009, 0.284877239027992, 0.941624935716391, 0.437050486216322, 0.127039125887677, 0.770586679922417, 0.432496567023918, 0.0606693513691425, 0.0242986304219812, 0.262240454787388, 0.740338553208858, 0.233443210599944, 0.221121123759076, 0.73616913985461, 0.674190784106031, 0.873344477498904, 0.723542936611921, 0.131069246213883, 0.79439337644726, 0.519966798601672, 0.266721702413633, 0.243282563285902, 0.629439793294296, 0.892523652641103, 0.148941732710227, 0.884265299653634, 0.340188351925462, 0.0987947126850486, 0.207283199066296, 0.517614830518141, 0.302085759118199, 0.292795462301001, 0.421875284751877, 0.417068249545991, 0.291902299504727, 0.710283993976191, 0.188545442651957, 0.0847026186529547, 0.805972188012674, 0.0475024497136474, 0.456646212609485, 0.140958706848323, 0.15486252354458, 0.853021269664168, 0.356763744493946, 0.0120942809153348, 0.69171148375608, 0.931538006057963, 0.153313032351434, 0.977238555438817, 0.728657926665619, 0.93528311746195, 0.211350154597312, 0.630144316004589, 0.593751852633432, 0.951586892362684, 0.0468777974601835, 0.242175498278812, 0.221412441693246, 0.769543117145076, 0.652123523876071, 0.980256935581565, 0.257640898227692, 0.913091017864645, 0.950794955715537, 0.0538449494633824, 0.61790313757956, 0.63308909162879, 0.422807751921937, 0.343203716911376, 0.642954273847863, 0.425245749298483, 0.477377250324935, 0.500695207854733, 0.583814869169146, 0.079744856338948, 0.874253290472552, 0.215712060686201, 0.758949395734817, 0.928820809116587, 0.245260093361139, 0.48929658671841, 0.581125998171046, 0.760571266058832, 0.428966633277014, 0.721965149044991, 0.427350679179654, 0.104427711339667, 0.718007773160934, 0.299680801108479, 0.02330700494349, 0.422856710851192, 0.307285945862532, 0.408713722834364, 0.238417038926855, 0.274429843295366, 0.222588079283014, 0.0605373333673924, 0.479100842494518, 0.23970361892134, 0.593597957631573, 0.424488271353766, 0.205221332842484, 0.00141813815571368, 0.722187550971285, 0.446109322831035, 0.625767398159951, 0.549309420632198, 0.472548026591539, 0.0211170339025557, 0.2452609075699, 0.674238374922425, 0.0462419209070504, 0.315698913764209, 0.967988208169118, 0.00191463530063629, 0.251422700239345, 0.0951290945522487, 0.813356038881466, 0.468223126139492, 0.672062526689842, 0.445931671652943, 0.192865697667003, 0.312888614833355, 0.517461549956352, 0.659990916261449, 0.040324438130483, 0.829509859904647, 0.827691956423223, 0.157019709004089, 0.782542092492804, 0.976525055011734, 0.614973713178188, 0.721110025420785, 0.931228280765936, 0.629504784243181, 0.0730282079894096, 0.456155133200809, 0.863284317078069, 0.867016285192221, 0.534874323289841, 0.9973312120419), actual = c(0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"