Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.7801828        0.5835776 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5835776        0.5835776        0.4799142 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.201456701848656, 0.219051119871438, 0.113347037695348, 0.609873606823385, 0.368924990529194, 0.29564376687631, 0.368258898844942, 0.9394555632025, 0.401204207912087, 0.919641204178333, 0.792284700786695, 0.736767801223323, 0.951236627297476, 0.804367191158235, 0.460791308665648, 0.66152432677336, 0.853349803946912, 0.344710598001257, 0.662472297437489, 0.458467320539057, 0.155964271631092, 0.701808483339846, 0.502460372634232, 0.416848902124912, 0.488567697117105, 0.934306883485988, 0.282773367129266, 0.2288183120545, 0.30358720314689, 0.764113783370703, 0.53136068279855, 0.391822967212647, 0.6234118307475, 0.540748755214736, 0.595628986833617, 0.734717083862051, 0.166021379176527, 0.202268772060052, 0.645473276032135, 0.217008350184187, 0.0706045886036009, 0.942157608922571, 0.398997193202376, 0.678179392358288, 0.209969113813713, 0.414102338720113, 0.649150470271707, 0.959733769064769, 0.302225965773687, 0.91439619846642, 0.546948106493801, 0.452341191004962, 0.60794569295831, 0.321607382502407, 0.383729089982808, 0.444244636455551, 0.624226301908493, 0.544663645094261, 0.20527814142406, 0.360320706153288, 0.040466494159773, 0.887690714560449, 0.417884465772659, 0.926682978868484, 0.284184137592092, 0.789400941226631, 0.82785958237946, 0.374149504583329, 0.384008155670017, 0.999584707897156, 0.762588125420734, 0.129101905738935, 0.877415666123852, 0.481938086682931, 0.370268200291321, 0.348243926186115, 0.641311360988766, 0.269063607091084, 0.0471288396511227, 0.968612790107727, 0.968045495683327, 0.196794274728745, 0.947112408000976, 0.765528028830886, 0.863026136066765, 0.832531911088154, 0.595742721576244, 0.638710601720959, 0.153487960109487, 0.499009915161878, 0.0360208260826766, 0.0414323552977294, 0.484792318660766, 0.00528221391141415, 0.892208392964676, 0.244076740462333, 0.532497123349458, 0.587150929961354, 0.694860220421106, 0.498985418817028, 0.762110700597987, 0.0149964413139969, 0.65949661587365, 0.302502018865198, 0.292785889934748, 0.351990298135206, 0.647743442794308, 0.273015098180622, 0.450302042299882, 0.114629267947748, 0.827359401620924, 0.653075022390112, 0.535477326717228, 0.582441185601056, 0.791175528196618, 0.293893308844417, 0.298809168860316, 0.661545257782564, 0.423667124006897, 0.124753899173811, 0.373512132558972, 0.886162069393322, 0.463026366895065, 0.151646211044863, 0.894378797151148, 0.227510339580476, 0.547997909365222, 0.508192837005481, 0.241339210886508, 0.240011381683871, 0.311649405397475, 0.968050508527085, 0.599096854915842, 0.637413760880008, 0.786051682662219, 0.437226230744272, 0.109318301081657, 0.904349416028708, 0.946054111002013, 0.0796612170524895, 0.399696453940123, 0.0871732772793621, 0.376998950028792, 0.48342532874085, 0.786264472641051, 0.775667964713648, 0.943888161564246, 0.952305457554758, 0.189752117032185, 0.420641569886357, 0.190479804528877, 0.989642997970805, 0.489150430541486, 0.084204853978008, 0.390134365763515, 0.0147925023920834, 0.470287885516882, 0.090696913190186, 0.135117741534486, 0.990759353153408, 0.779008375480771, 0.194568160455674, 0.389377607032657, 0.291762117529288, 0.00643355725333095, 0.516629092162475, 0.41648188745603, 0.658683406887576, 0.108200255548581, 0.768242413410917, 0.190544702811167, 0.263688158709556, 0.729339007986709, 0.172114968998358, 0.149103843374178, 0.440565862925723, 0.833281547296792, 0.695218272740021, 0.390085279941559, 0.339413010515273, 0.498162737116218, 0.0423450313974172, 0.0098340071272105, 0.41285762283951, 0.0666962403338403, 0.958409780170768, 0.298174078343436, 0.256246066652238, 0.0658257075119764, 0.87876049708575, 0.482621154515073, 0.216624915832654, 0.44878518790938, 0.088363291695714, 0.457731769187376, 0.768381537171081, 0.766242709942162, 0.727117834612727, 0.495207319036126, 0.421806604368612, 0.386621554847807, 0.430656284326687, 0.0285635909531265, 0.474725995911285, 0.536172985099256, 0.271581876091659, 0.972668351605535, 0.193099688738585, 0.0847379732877016, 0.736016831360757, 0.303687639767304, 0.650862160371616, 0.0848548379726708, 0.77694200607948, 0.642528876895085, 0.0959574040025473, 0.286727264756337, 0.866959014907479, 0.454692015890032, 0.096737390384078, 0.322635099757463, 0.579038211610168, 0.580152037320659, 0.898936544079334, 0.52554197399877, 0.186645373003557, 0.0881535534281284, 0.904701249906793, 0.130270675988868, 0.939770214026794, 0.247681420762092, 0.861407934920862, 0.441273916745558, 0.355543807148933, 0.0335267814807594, 0.826373571995646, 0.899931669933721, 0.173888030461967, 0.316307598957792, 0.320930951042101, 0.339963583042845, 0.0787352412007749, 0.251287758350372, 0.77994687971659, 0.856780105503276, 0.896577283274382, 0.563024773728102, 0.22407366707921, 0.307268842123449, 0.712102404329926, 0.143745381850749, 0.42842988925986, 0.241670283954591, 0.961172390263528, 0.0495496918447316, 0.614433483686298, 0.789373262785375, 0.281709356699139, 0.141406260896474, 0.301288994029164, 0.439974560635164, 0.0130441095679998, 0.825395328458399, 0.393630845006555, 0.654226690065116, 0.885628430871293, 0.223956804024056, 0.35273456084542, 0.826925416942686, 0.69029522780329, 0.638352665351704, 0.0230738252867013, 0.316837638616562, 0.740637073293328, 0.274985735537484, 0.53099312982522, 0.351396087091416, 0.657743650255725, 0.587882083840668, 0.196035849163309, 0.0148253969382495, 0.345420632511377, 0.812992190709338, 0.865403312258422, 0.582665971014649, 0.82180930650793, 0.167803856544197, 0.796422425657511, 0.371611512964591, 0.993842525174841, 0.41436720197089, 0.205489430110902, 0.716733672656119, 0.458822620101273, 0.578599828761071, 0.607687854208052, 0.693637580377981, 0.750742949778214, 0.995724918087944, 0.714478724403307, 0.132450151722878, 0.183209860231727, 0.467771284049377, 0.0826961817219853, 0.14186054142192, 0.793024797923863, 0.51911713113077, 0.489402228500694, 0.0759494996163994, 0.859137462452054, 0.4528208442498, 0.932843324029818, 0.706815020879731, 0.698775956872851, 0.0513827737886459, 0.0161763029173017, 0.530523683642969, 0.461726625449955, 0.256306912982836, 0.00477365939877927, 0.266056811902672, 0.627469113329425, 0.862108947476372, 0.498983671655878, 0.455078031402081, 0.334197213174775, 0.425887315534055, 0.364737576805055, 0.950122905429453, 0.123888172442093, 0.996022067731246, 0.737263809423894, 0.00646091508679092, 0.60520950332284, 0.747414084849879, 0.867966360412538, 0.252403327263892, 0.158551021013409, 0.238577372394502, 0.0364445648156106, 0.00777287012897432, 0.343286595772952, 0.697291863849387, 0.331807109061629, 0.200583049794659, 0.226705939043313, 0.45846631960012, 0.791838334407657, 0.459464798914269, 0.724006848642603, 0.936585700372234, 0.363054760964587, 0.628687757067382, 0.20423406874761, 0.975135315442458, 0.552867567399517, 0.230652822880074, 0.628859275719151, 0.870046926429495, 0.603107450297102, 0.278907448984683, 0.352349304128438, 0.160513746319339, 0.152179532218724, 0.875331811374053, 0.225095381261781, 0.808360312366858, 0.811617432394996, 0.282426465069875, 0.0788475784938782, 0.409952118527144, 0.383109123446047, 0.515473572071642, 0.412721624365076, 0.218415842624381, 0.729791731573641, 0.9396016667597, 0.684426181018353, 0.876059743808582, 0.981834918493405, 0.564047290943563, 0.0934421131387353, 0.218408719403669, 0.00822576833888888, 0.438270804937929, 0.152434007730335, 0.756793439621106, 0.0335056106559932, 0.601477819494903, 0.504893111530691, 0.401335422648117, 0.344602584606037, 0.800247204722837, 0.316424045013264, 0.0589709130581468, 0.782772632315755, 0.946377795888111, 0.618483996717259, 0.186440889257938, 0.604438232956454, 0.852828418137506, 0.234194854740053, 0.632142305839807, 0.613133736187592, 0.571732656797394, 0.927256806287915, 0.473670367151499, 0.276175931096077, 0.0850141288246959, 0.953279553679749, 0.797700002091005, 0.0974117023870349, 0.507719955639914, 0.704517665551975, 0.686722739599645, 0.620308642508462, 0.81387992715463, 0.428880700841546, 0.186871261801571, 0.348677397705615, 0.713751329807565, 0.462417509872466, 0.400464226026088, 0.906875351676717, 0.99288085126318, 0.772981346584857, 0.0164370983839035, 0.970306506613269, 0.685083504999056, 0.232609511120245, 0.404860647628084, 0.0910278563387692, 0.129706740844995, 0.437454010127112, 0.0550687212962657, 0.810091499472037, 0.412982208421454, 0.0810239675920457, 0.550183262210339, 0.281987360911444, 0.744110094383359, 0.0955866859294474, 0.483185780001804, 0.479442035313696, 0.961110337404534, 0.258148958208039, 0.395119450287893, 0.381210901774466, 0.313893217826262, 0.777492589084432, 0.53714389167726, 0.163952135713771, 0.33652776828967, 0.735388391418383, 0.17236564680934, 0.680955602321774, 0.652210694970563, 0.847148152301088, 0.216432612622157, 0.627772771753371, 0.186347708804533, 0.325390723533928, 0.143109238473698, 0.924810939701274, 0.0686209867708385, 0.848941870266572, 0.79704024665989, 0.0807560652028769, 0.0707749158609658, 0.26035936945118, 0.218190045328811, 0.223629345884547, 0.262987583875656, 0.64504866534844, 0.531379155116156, 0.0393772018142045, 0.0690809118095785, 0.388422133401036, 0.428608767222613, 0.499928384786472, 0.343375506112352, 0.374588372185826, 0.94797076866962, 0.0133822953794152, 0.775901418412104, 0.696787410415709, 0.371460482478142, 0.487657648976892, 0.354963338933885, 0.785684489645064, 0.966922180727124, 0.191722710151225, 0.609005209524184, 0.891312105581164, 0.923203069949523, 0.105851531494409, 0.533962194342166, 0.0488976463675499, 0.456334114074707, 0.235155896749347, 0.484030243707821, 0.308077871333808, 0.695083116181195, 0.367196292849258, 0.538693580776453, 0.511203471105546, 0.220187788130715, 0.245975320925936, 0.744413715787232, 0.00671676406636834, 0.205438049277291, 0.232601548545063, 0.632735966937616, 0.615982637740672, 0.210598361445591, 0.490240333834663, 0.0994806592352688, 0.651534458855167, 0.980161278508604, 0.504432093817741, 0.635652631288394, 0.100127476034686, 0.553575203055516, 0.918097687186673, 0.90101232053712, 0.718816217035055, 0.841497197048739, 0.925389665411785, 0.497388126561418, 0.243249156977981, 0.91557330545038, 0.911559984320775, 0.168140349211171, 0.476743252947927, 0.123354934854433, 0.833213384263217, 0.912956011481583, 0.714614542899653, 0.00648557231761515, 0.226796677103266, 0.994424465345219, 0.962745151482522, 0.505476031918079, 0.547655599191785, 0.399888063315302, 0.941940451040864, 0.852149308193475, 0.439688819460571, 0.903538586804643, 0.653444170951843, 0.545747802127153, 0.453325033653527, 0.678099746350199, 0.779583371477202, 0.957796493778005, 0.917229504091665, 0.790693296585232, 0.435504353139549, 0.969724306836724, 0.0452585341408849, 0.460113180801272, 0.825783068779856, 0.520150609547272, 0.531916292151436, 0.384804990142584, 0.778509094845504, 0.257014006143436, 0.167111799586564, 0.938662316882983, 0.404209463624284, 0.0689678508788347, 0.900458045070991, 0.774157781619579, 0.0406335638836026, 0.816805241629481, 0.283097743522376, 0.0718282398302108, 0.427879708120599, 0.802040670532733, 0.570029260590672, 0.707770112901926, 0.620808940846473, 0.433062255615368, 0.277063306188211, 0.279823633376509, 0.998200650792569, 0.186041741166264, 0.929010110441595, 0.870326915523037, 0.997991375625134, 0.832120267208666, 0.357840240001678, 0.974660083418712, 0.384553051553667, 0.938428963301703, 0.827321944059804, 0.871448290999979, 0.67802526964806, 0.263804465997964, 0.796647425508127, 0.748855635290965, 0.749509007437155, 0.0192688936367631, 0.32785074156709, 0.140565141104162, 0.416157419094816, 0.087335345800966, 0.0567939730826765, 0.744339808356017, 0.455203930148855, 0.470812253188342, 0.188279427355155, 0.951204287121072, 0.600820749998093, 0.439568229019642, 0.459254611516371, 0.342269963817671, 0.5829417349305, 0.529025907861069, 0.831184732727706, 0.238895102636889, 0.258215134497732, 0.817754362244159, 0.665672780713066, 0.0403843619860709, 0.322637806879357, 0.918288704473525, 0.794758048374206, 0.519579455256462, 0.0275581160094589, 0.76992299943231, 0.707223821664229, 0.765714680543169, 0.579749602358788, 0.21860490902327, 0.510431329254061, 0.369768033036962, 0.753112158970907, 0.728772242320701, 0.835373499663547, 0.772417181869969, 0.213300045812503, 0.444329440593719, 0.205501375719905, 0.732403170783073, 0.566753115737811, 0.48762011481449, 0.188445140840486, 0.56142038712278, 0.993109114002436, 0.293471257900819, 0.680828281678259, 0.166503153275698, 0.231650333385915, 0.928925290005282, 0.423994528129697, 0.607558090938255, 0.222721664933488, 0.346109833102673, 0.315208598738536, 0.97133128112182, 0.593821821035817, 0.0295708270277828, 0.216196818742901, 0.671126988483593, 0.776470326120034, 0.460915389936417, 0.982919324422255, 0.301740978611633, 0.56937594502233, 0.210943315178156, 0.683629957260564, 0.334626150550321, 0.444830486085266, 0.101043055066839, 0.844715142156929, 0.839268740033731, 0.0694898010697216, 0.121016520541161, 0.629299964755774, 0.964529492426664, 0.625160885741934, 0.0611155705992132, 0.00648582889698446, 0.67870039306581, 0.121515229810029, 0.372295242734253, 0.108922889921814, 0.614244650583714, 0.935851539950818, 0.367207715986297, 0.551686696242541, 0.0225134191568941, 0.842678752494976, 0.801407334394753, 0.00910445069894195, 0.698202532250434, 0.572749466635287, 0.177381451940164, 0.730324385687709, 0.0849728749599308, 0.21492412686348, 0.0714165854733437, 0.317762901540846, 0.227204893250018, 0.186903026187792, 0.954572671325877, 0.83487212494947, 0.537910902872682, 0.279641026398167, 0.442384544527158, 0.206546983681619, 0.324554480379447, 0.766387399984524, 0.0393977081403136, 0.993384808767587, 0.726247837767005, 0.537773099495098, 0.649380387505516, 0.385259898379445, 0.983198652742431, 0.804411935387179, 0.101038476219401, 0.789230612106621, 0.00331063778139651, 0.795017978409305, 0.530446070479229, 0.630562277976424, 0.624686856055632, 0.128289368934929, 0.9715533235576, 0.310170844430104, 0.146818578243256, 0.337680987082422, 0.0775533842388541, 0.362893872195855, 0.785084374947473, 0.270943327108398, 0.807527336524799, 0.150444021681324, 0.67081922153011, 0.951783329248428, 0.320852796314284, 0.653194396756589, 0.836611107457429, 0.246816425118595, 0.239939476363361, 0.600717251654714, 0.341231076978147, 0.111935669789091, 0.938143700594082, 0.901535075856373, 0.850765081122518, 0.636433457490057, 0.306425133254379, 0.318019391503185, 0.54712848062627, 0.61487511312589, 0.160898213041946, 0.356170500162989, 0.79366237251088, 0.971760937012732, 0.447676220210269, 0.790835399413481, 0.625628116773441, 0.0658155055716634, 0.499907673802227, 0.36871092999354, 0.490862803300843, 0.975192686310038, 0.680529720615596, 0.0476100915111601, 0.872445463668555, 0.149050114909187, 0.219657935434952, 0.353508508764207, 0.502346087014303, 0.945158133748919, 0.772673877421767, 0.224819606170058, 0.350207055918872, 0.732012086315081, 0.873987063067034, 0.689457493135706, 0.636835275683552, 0.240372100146487, 0.368265276541933, 0.433556131785735, 0.75435494328849, 0.544008712051436, 0.298135647084564, 0.819265518104658, 0.665101830847561, 0.493957689264789, 0.0620879330672324, 0.681060443399474, 0.862761680502445, 0.324985092738643, 0.977215108927339, 0.823404669761658, 0.540688108885661, 0.502734223613515, 0.744231652235612, 0.156596370972693, 0.872686651768163, 0.423628131626174, 0.400671566370875, 0.917775263544172, 0.77138610323891, 0.681808377383277, 0.452554320450872, 0.252538002561778, 0.11436523986049, 0.362507057376206, 0.447217155015096, 0.883704692590982, 0.396478250157088, 0.738349972991273, 0.991714569274336, 0.747553765308112, 0.654025652213022, 0.288416256196797, 0.250463319709525, 0.714452264597639, 0.517054115189239, 0.969919657101855, 0.39802357996814, 0.595903495093808, 0.834685804555193, 0.544292877893895, 0.41297980141826, 0.200037813745439, 0.370275025023147, 0.600590347778052, 0.609610063256696, 0.626688497141004, 0.56080376310274, 0.755517841083929, 0.792169836815447, 0.26212794566527, 0.074259570101276, 0.534627625253052, 0.231602278305218, 0.245393007062376, 0.230524283833802, 0.557645610067993, 0.481101598590612, 0.773013222264126, 0.929191084578633, 0.22511565242894, 0.851445480482653, 0.463843068340793, 0.0374214395415038, 0.715689422329888, 0.601694390876219, 0.969791184645146, 0.976712763076648, 0.7016118937172, 0.766688707750291, 0.472514446591958, 0.36551953246817, 0.331264414358884, 0.992437077919021, 0.320031621260568, 0.950125572737306, 0.717766631860286, 0.813880926696584, 0.32696903962642, 0.301412432221696, 0.177668193820864, 0.112445318605751, 0.429607350612059, 0.959678318817168, 0.928505709161982, 0.699119550175965, 0.0290164262987673, 0.237190922955051, 0.485756105743349, 0.77670069038868, 0.463191956514493, 0.962684673722833, 0.271659363526851, 0.369840914383531, 0.335254119243473, 0.396904458058998, 0.49891025503166, 0.769973821472377, 0.625505808042362, 0.856775793014094, 0.722888796590269, 0.715953588951379, 0.233971412526444, 0.597770442487672, 0.535214219940826, 0.118101629428566, 0.534327375935391, 0.309616683283821, 0.551206413656473, 0.964317113393918, 0.373434247681871, 0.703752365894616, 0.938141644233838, 0.0523265209048986, 0.227025441592559, 0.134989511920139, 0.123406356666237, 0.398882368113846, 0.278838219121099, 0.776619736570865, 0.368628847878426, 0.593633727869019, 0.700206863693893, 0.25474389619194, 0.155629312153906, 0.934165432816371, 0.991837139939889, 0.057447409722954, 0.943795511499047, 0.272855797316879, 0.882493691053241, 0.143199838930741, 0.95718794548884, 0.693870160263032, 0.79677801951766, 0.61028751777485, 0.753952677594498, 0.590478969970718, 0.951689760200679, 0.0277654409874231, 0.226584370015189, 0.41523614525795, 0.026092471787706, 0.270383585244417, 0.856513863662258, 0.891346230404451, 0.30953812552616, 0.914728746050969, 0.486091797472909, 0.30640684789978, 0.863732418278232, 0.928560137283057, 0.232640993082896, 0.298156645148993, 0.0245047779753804, 0.392493335064501, 0.424653824418783, 0.355458157602698, 0.398945016553625, 0.199119812808931, 0.21475188434124, 0.884346815291792, 0.866867242613807, 0.643868694780394, 0.0648347064852715, 0.278204016154632, 0.208940691314638, 0.693238066975027, 0.191341656027362, 0.74134621815756, 0.852487914729863, 0.503602144308388, 0.229604961816221, 0.166401085909456, 0.576330305542797, 0.886555701494217, 0.280272559961304, 0.394702593097463, 0.494960801675916, 0.809226805577055, 0.878338831244037, 0.105579645605758, 0.102355383336544, 0.646173293935135, 0.490309946704656, 0.957789342151955, 0.381254955427721, 0.149251641239971, 0.573525125626475, 0.401716575026512, 0.849541595205665, 0.541371488012373, 0.135136995464563, 0.144782077288255, 0.713669284246862, 0.0672526506241411, 0.186731733847409, 0.984409082215279, 0.859074807260185, 0.231707681203261, 0.261040561133996, 0.867515134857967, 0.592392103979364, 0.189096163958311, 0.773300814675167, 0.548576541244984, 0.656945610884577, 0.9778970007319, 0.641099410830066, 0.62877728859894, 0.141378773842007), actual = c(1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.0411798190325499, 0.523413356859237, 0.668195620412007, 0.952263367827982, 0.178391861962155, 0.827997614163905, 0.645356448367238, 0.973946369253099, 0.132493383716792, 0.740161378402263, 0.92154569295235, 0.641407434595749, 0.153577671851963, 0.725094196619466, 0.313277590554208, 0.164227596949786, 0.893047761172056, 0.752508820500225, 0.315371954115108, 0.348233954748139, 0.762177523691207, 0.0483796282205731, 0.274130225647241, 0.443131692707539, 0.00101371551863849, 0.256032469449565, 0.275114183314145, 0.142028328962624, 0.812711326871067, 0.912460276624188, 0.10326205776073, 0.841427990002558, 0.61126780253835, 0.329282797407359, 0.142239785986021, 0.821275070309639, 0.780661690514535, 0.26190235116519, 0.211604643147439, 0.62861054437235, 0.464069680776447, 0.322264323942363, 0.728678891900927, 0.906816948438063, 0.783918980509043, 0.545600533951074, 0.0701004513539374, 0.871474091196433, 0.424396481830627, 0.0785410441458225, 0.103498157579452, 0.24492522305809, 0.603962374851108, 0.382595685077831, 0.50863016792573, 0.175339735811576, 0.658274138113484, 0.203501020558178, 0.305434583919123, 0.432573636993766, 0.707771575078368, 0.906022465089336, 0.14694298710674, 0.178964246297255, 0.208800284890458, 0.0771448106970638, 0.508655076147988, 0.367073874920607, 0.592184603214264, 0.707996093202382, 0.281665215734392, 0.366585279116407, 0.173129781382158, 0.33575068693608, 0.791260708356276, 0.491474500391632, 0.720214576926082, 0.420279128942639, 0.929501743754372, 0.19041685666889, 0.394294309429824, 0.555254014907405, 0.631293169455603, 0.194923343835399, 0.863987488904968, 0.903606995940208, 0.494789325864986, 0.562926441896707, 0.871988346334547, 0.255076414206997, 0.785742417676374, 0.802771270042285, 0.95976178208366, 0.62477742577903, 0.934698603814468, 0.89081133948639, 0.643165507121012, 0.582936644786969, 0.935036689974368, 0.0857103269081563, 0.867930937325582, 0.402441277401522, 0.655727747827768, 0.850900107994676, 0.732301668263972, 0.508615735685453, 0.968940861755982, 0.899906602222472, 0.580295152962208, 0.22335182223469, 0.0209852193947881, 0.396956542041153, 0.905223330017179, 0.196723274420947, 0.152746594976634, 0.898508043959737, 0.403902509715408, 0.303424558602273, 0.618619073880836, 0.891277143498883, 0.636504473397508, 0.126430509379134, 0.647466330323368, 0.395041783805937, 0.24131315969862, 0.109482451574877, 0.844789973925799, 0.0584531840868294, 0.78947232128121, 0.530293501215056, 0.00260159326717257, 0.407371293287724, 0.238590801600367, 0.480457810917869, 0.608511617872864, 0.904840021627024, 0.939662706572562, 0.123220288893208, 0.365545069100335, 0.892849762458354, 0.284329424146563, 0.801462507573888, 0.748668718617409, 0.174462827388197, 0.447270514676347, 0.113841783022508, 0.544056014390662, 0.53994684969075, 0.730774954427034, 0.503333078697324, 0.295749038690701, 0.969524247338995, 0.434395742835477, 0.842684973962605, 0.649972719373181, 0.575980350375175, 0.882844649255276, 0.0865197530947626, 0.84262306895107, 0.659482959192246, 0.0719167583156377, 0.806196700083092, 0.650481903227046, 0.217753190780059, 0.937342984834686, 0.293686231132597, 0.239616169128567, 0.770879278192297, 0.418879755074158, 0.905240705935284, 0.260578805115074, 0.336639801273122, 0.819653241895139, 0.746450604172423, 0.27264480246231, 0.155928393825889, 0.995213754242286, 0.365025124745443, 0.127613751217723, 0.227009742055088, 0.292377897072583, 0.293584022670984, 0.641722678905353, 0.620290004881099, 0.322078725323081, 0.276335437549278, 0.964971010806039, 0.848287266911939, 0.929575266782194, 0.47666884562932, 0.798303826944903, 0.887446891050786, 0.800652540521696, 0.330641962587833, 0.749081074027345, 0.781532554188743, 0.307882678462192, 0.789279964752495, 0.401035614078864, 0.429212532471865, 0.00944905960932374, 0.336044769035652, 0.431302116019651, 0.241380653111264, 0.503860829398036, 0.651697007240728, 0.966081848135218, 0.0143685180228204, 0.340243828250095, 0.542580901412293, 0.275959914084524, 0.176543432986364, 0.49851951841265, 0.268440344603732, 0.918684236705303, 0.0980747730936855, 0.134827275993302, 0.0509128370322287, 0.737034593941644, 0.604621943086386, 0.59990528668277, 0.369996239896864, 0.491217470029369, 0.831908592954278, 0.715267927385867, 0.496036528609693, 0.907369080698118, 0.74638829450123, 0.211771982489154, 0.00703710946254432, 0.886910097207874, 0.604504317976534, 0.240895943483338, 0.90277739497833, 0.403701535658911, 0.977322598686442, 0.754251671023667, 0.554999499116093, 0.702389638638124, 0.514191176975146, 0.999293497297913, 0.392738428432494, 0.00897265179082751, 0.831326244398952, 0.966601563617587, 0.0677273909095675, 0.170017640804872, 0.64855427807197, 0.778927648440003, 0.631833255989477, 0.0042268168181181, 0.148402212420478, 0.773000611923635, 0.438920577755198, 0.530285842716694, 0.655883646337315, 0.437860404839739, 0.890749665210024, 0.128806235035881, 0.942777961259708, 0.444190746871755, 0.220431528519839, 0.726401128573343, 0.800855285255238, 0.00779947941191494, 0.100837137550116, 0.0376549209468067, 0.716805185656995, 0.518559378804639, 0.86943828384392, 0.249849310377613, 0.529629450757056, 0.573878572322428, 0.471464906586334, 0.0399241424165666, 0.438710534246638, 0.614912206539884, 0.450994197744876, 0.973757583415136, 0.570715352892876, 0.0167254356201738, 0.964818396372721, 0.506278933491558, 0.492167932214215, 0.402110422495753, 0.811189410509542, 0.78334752516821, 0.970072969328612, 0.904133279807866, 0.17324258829467, 0.583815762540326, 0.568542235763744, 0.107383678667247, 0.0755745451897383, 0.241104107350111, 0.101628694217652, 0.627830286277458, 0.256583385867998, 0.289574480848387, 0.377440622076392, 0.812563720392063, 0.567953921854496, 0.452300617005676, 0.478758650831878, 0.391364448703825, 0.475348717300221, 0.589404075872153, 0.981634134426713, 0.367903125938028, 0.89280684874393, 0.0581583422608674, 0.0219576368108392, 0.0191046928521246, 0.235833945451304, 0.733761283801869, 0.333523924928159, 0.790838544489816, 0.915203798096627, 0.939622406382114, 0.0551175016444176, 0.0171773419715464, 0.218894426943734, 0.233071158872917, 0.215455287601799, 0.25175525364466, 0.694608576362953, 0.663210454629734, 0.942061909474432, 0.400005543837324, 0.519642534200102, 0.383690954651684, 0.824757691239938, 0.0689538877923042, 0.210082851350307, 0.486942230025306, 0.339500684989616, 0.79331246460788, 0.966893937205896, 0.827472180593759, 0.49320302833803, 0.443218149011955, 0.912726016948, 0.988281239755452, 0.895611737854779, 0.0854730892460793, 0.842532213544473, 0.968100072117522, 0.283734967000782, 0.945026602596045, 0.990577954566106, 0.854958558920771, 0.859099340392277, 0.734122721478343, 0.166121001821011, 0.2416969507467, 0.522055494133383, 0.630464036017656, 0.735717586707324, 0.437826770590618, 0.393428578274325, 0.75342203816399, 0.444418816128746, 0.0588469314388931, 0.153120866511017, 0.312788656447083, 0.389603982213885, 0.15876150294207, 0.919200825272128, 0.246366294799373, 0.115330117056146, 0.405438492773101, 0.96518496517092, 0.113404529634863, 0.049196673091501, 0.364493142347783, 0.245327842188999, 0.665987831307575, 0.571294830180705, 0.394972101319581, 0.291990682017058, 0.585424792254344, 0.63115872675553, 0.309193543158472, 0.0156662161462009, 0.25726392166689, 0.751997565152124, 0.059302257373929, 0.699369882931933, 0.982544145546854, 0.351064437767491, 0.357034098356962, 0.96235391148366, 0.813561018323526, 0.440413910662755, 0.988820164231583, 0.488375776447356, 0.036547232652083, 0.559017928084359, 0.045873146969825, 0.796595342224464, 0.903452565195039, 0.0336635671555996, 0.760274360422045, 0.314318183809519, 0.72850623819977, 0.747938323533162, 0.525352246826515, 0.702705509960651, 0.311839600559324, 0.513391194166616, 0.824130123946816, 0.0664325510151684, 0.114046368282288, 0.584667910123244, 0.677437151782215, 0.676604184089229, 0.483648788183928, 0.164598981384188, 0.64531801152043, 0.909936375217512, 0.313820382580161, 0.575225132051855, 0.0548165822401643, 0.642323896987364, 0.038228313671425, 0.597641501808539, 0.0772980723995715, 0.830585824325681, 0.861579426564276, 0.161025305977091, 0.755932049360126, 0.931557278614491, 0.809492717031389, 0.275533576263115, 0.0380868064239621, 0.469029309460893, 0.261657998664305, 0.656897934386507, 0.66382393008098, 0.985909598181024, 0.687821210594848, 0.170541926985607, 0.671897353371605, 0.900708568049595, 0.0218796776607633, 0.400136388838291, 0.022367222700268, 0.26655293838121, 0.197078244993463, 0.367807811358944, 0.166825590422377, 0.383136352989823, 0.88049743627198, 0.300816923612729, 0.602735764812678, 0.190817771246657, 0.648408461594954, 0.849215358495712, 0.846831531263888, 0.476308340439573, 0.900115475989878, 0.585012084338814, 0.105478556826711, 0.992212783778086, 0.949218258494511, 0.864894627127796, 0.202324742916971, 0.426031113602221, 0.212676962371916, 0.419081838801503, 0.13871008856222, 0.230799533426762, 0.169087269576266, 0.506735685281456, 0.759031276684254, 0.536020858446136, 0.715565304737538, 0.491743465187028, 0.912828758358955, 0.193429308477789, 0.957930399803445, 0.126576442271471, 0.793025129940361, 0.0327943908050656, 0.677435625111684, 0.322989606996998, 0.197471833322197, 0.609212658368051, 0.316046159015968, 0.907519346103072, 0.352710221661255, 0.867030094610527, 0.529761503217742, 0.965209882706404, 0.123369409702718, 0.40460418513976, 0.572088656947017, 0.764707412570715, 0.529961882159114, 0.147640263661742, 0.37833660421893, 0.181970193050802, 0.127286605304107, 0.359641184331849, 0.899223451968282, 0.773644647793844, 0.676683519501239, 0.865037936251611, 0.809131226269528, 0.126011986052617, 0.819052064791322, 0.207538286689669, 0.545422197319567, 0.295288719236851, 0.755304214777425, 0.502133809262887, 0.648549086181447, 0.594653379172087, 0.977141427574679, 0.17251782421954, 0.242359617725015, 0.907023558858782, 0.480022230418399, 0.511667464859784, 0.646818860899657, 0.422702569980174, 0.405562488362193, 0.639874979620799, 0.184849763987586, 0.62522346386686, 0.275502608157694, 0.0267596961930394, 0.383074595127255, 0.986746487207711, 0.516569935716689, 0.288169409381226, 0.624091062461957, 0.955773489084095, 0.481431865133345, 0.63107846560888, 0.333036449737847, 0.739300275454298, 0.0470200758427382, 0.675360394408926, 0.587864383356646, 0.698122635716572, 0.662333589280024, 0.476004937430844, 0.256070161471143, 0.345903945388272, 0.279916948871687, 0.615207155235112, 0.237544875359163, 0.648597738007084, 0.877239298308268, 0.934712336398661, 0.5765413234476, 0.653015058254823, 0.612967988476157, 0.353434503078461, 0.221365528879687, 0.255714080994949, 0.910187331028283, 0.662172140320763, 0.0750072153750807, 0.139332434860989, 0.0047444817610085, 0.0801541560795158, 0.479716228554025, 0.717356371693313, 0.656693904194981, 0.34872041689232, 0.179643519921228, 0.372962414985523, 0.914037242997438, 0.499099761480466, 0.023176226997748, 0.975843058666214, 0.823025931837037, 0.833318936172873, 0.815942455548793, 0.338010631734505, 0.127358203288168, 0.543931448832154, 0.71907499874942, 0.344645586330444, 0.300024725496769, 0.649966524215415, 0.928046339424327, 0.482679675333202, 0.750681683421135, 0.302086894633248, 0.290909475414082, 0.973392799729481, 0.479699337622151, 0.656899865483865, 0.792568664997816, 0.245726102730259, 0.0730438819155097, 0.936145647894591, 0.193336632568389, 0.629444354912266, 0.146325158653781, 0.195209020283073, 0.638965689111501, 0.846527272369713, 0.640006583649665, 0.825801651226357, 0.735649993410334, 0.361970651196316, 0.344912564381957, 0.244917370378971, 0.889824999030679, 0.679762061685324, 0.14945706538856, 0.289911134634167, 0.0369964980054647, 0.960425061872229, 0.514533822191879, 0.51244045747444, 0.558710080338642, 0.94424228835851, 0.830897193634883, 0.101109054405242, 0.99352576257661, 0.161725076846778, 0.0598223120905459, 0.737818622263148, 0.959304058225825, 0.609479570761323, 0.579262104351074, 0.207181419013068, 0.692480818834156, 0.698773942422122, 0.496980896219611, 0.885241077281535, 0.216494007967412, 0.893507578177378, 0.000554139725863934, 0.943760242080316, 0.0183970779180527, 0.904218125157058, 0.91521922359243, 0.169274569023401, 0.293938745744526, 0.664390081074089, 0.0323213371448219, 0.793491165386513, 0.418187509523705, 0.921632860787213, 0.371574083110318, 0.573882615426555, 0.56952243950218, 0.264433623058721, 0.421366090420634, 0.911187107674778, 0.728825299534947, 0.261277145706117, 0.443427053745836, 0.817834192421287, 0.708881997969002, 0.197212925413623, 0.316282172920182, 0.147181282984093, 0.350642765406519, 0.250721560092643, 0.192388291470706, 0.00641451496630907, 0.374441079329699, 0.207998023135588, 0.23767550336197, 0.0430998848751187, 0.362178630428389, 0.741126508684829, 0.623432660708204, 0.325322451069951, 0.912983073387295, 0.0435371110215783, 0.0684796378482133, 0.140117014292628, 0.430914766155183, 0.327748076291755, 0.427721883170307, 0.298721513245255, 0.63699392369017, 0.360434562666342, 0.334255795227364, 0.946245790226385, 0.796921418746933, 0.311112995492294, 0.899307701969519, 0.438847101060674, 0.999767557485029, 0.868273593252525, 0.809031933778897, 0.0757934632711112, 0.340656030690297, 0.897916045738384, 0.183745933929458, 0.743851850973442, 0.977179158013314, 0.392091667279601, 0.783373794518411, 0.693283364875242, 0.623903260100633, 0.896892389515415, 0.453921576496214, 0.731624476145953, 0.398024168564007, 0.0703438092023134, 0.228365248767659, 0.395955461775884, 0.605450965231284, 0.231864460511133, 0.153759907931089, 0.773411840898916, 0.623038667486981, 0.899439177941531, 0.272979663684964, 0.303510646335781, 0.21430706512183, 0.89971789275296, 0.500057399505749, 0.812839567894116, 0.82775274082087, 0.12545890500769, 0.241138861281797, 0.570601109648123, 0.766602754825726, 0.173558844486251, 0.757148986682296, 0.924418356735259, 0.833249018294737, 0.549521256703883, 0.84267542255111, 0.06794074899517, 0.0116197941824794, 0.531765337567776, 0.70515041379258, 0.317795458249748, 0.652733347145841, 0.710262282751501, 0.803474079817533, 0.00159991392865777, 0.168056193506345, 0.222693223040551, 0.978308893041685, 0.446504026884213, 0.0920165840070695, 0.274387358687818, 0.99602014455013, 0.486141238361597, 0.307982749305665, 0.181752968113869, 0.399910122854635, 0.597879551816732, 0.502322665415704, 0.446060207905248, 0.795122141717002, 0.632754132151604, 0.782991321058944, 0.761133500607684, 0.163948526838794, 0.250174118438736, 0.701906777918339, 0.387938716216013, 0.247513329144567, 0.565127093810588, 0.660829307511449, 0.71117375488393, 0.050892231753096, 0.542980888858438, 0.947841014713049, 0.667434443486854, 0.953344095265493, 0.83820096263662, 0.689882080070674, 0.0864935892168432, 0.170980557100847, 0.75746526545845, 0.760969767114148, 0.839444191195071, 0.483863978646696, 0.292933164862916, 0.451262183021754, 0.227529333438724, 0.0230608547572047, 0.591592110460624, 0.336985654430464, 0.82569549116306, 0.577889422886074, 0.800575274741277, 0.1847070439253, 0.746679146541283, 0.0950275480281562, 0.0139385506045073, 0.142655872739851, 0.640019069658592, 0.962776920758188, 0.327675385866314, 0.534646577900276, 0.25585218030028, 0.285018023801968, 0.822862299392, 0.254284104332328, 0.368456488940865, 0.19658094109036, 0.653585216961801, 0.44076660531573, 0.628881092416123, 0.122120648855343, 0.577363274758682, 0.568100733915344, 0.409065390005708, 0.306561943143606, 0.376951891696081, 0.866497023031116, 0.0536021387670189, 0.0643037208355963, 0.414049964398146, 0.3507429221645, 0.698125312337652, 0.399354677647352, 0.483696135692298, 0.817952875280753, 0.200047267600894, 0.136149329366162, 0.925760475918651, 0.708399569848552, 0.699597349390388, 0.332209196407348, 0.00450082030147314, 0.94034707848914, 0.974198924377561, 0.853963975328952, 0.917649138951674, 0.170872601913288, 0.367938358802348, 0.4261901199352, 0.0670666538644582, 0.197673060698435, 0.673453980823979, 0.670596313662827, 0.286197098204866, 0.959079033927992, 0.339717640774325, 0.808500021928921, 0.634677219437435, 0.365720226429403, 0.445335543481633, 0.414993837242946, 0.842156484955922, 0.845735743874684, 0.390010891715065, 0.112285489216447, 0.759150060359389, 0.325106658041477, 0.174056680407375, 0.690685330191627, 0.519142247503623, 0.607874288689345, 0.347520547918975, 0.0136693699751049, 0.730217623291537, 0.655280358390883, 0.258320177672431, 0.200781762367114, 0.555311343632638, 0.183336750138551, 0.346862842561677, 0.820414871908724, 0.945713849039748, 0.202730202348903, 0.227757676271722, 0.572354795178398, 0.657783528091386, 0.821306709432974, 0.660084468778223, 0.697232679463923, 0.502136955969036, 0.144012234406546, 0.786068515386432, 0.302047160686925, 0.842079013586044, 0.331309332279488, 0.699520780704916, 0.493137976620346, 0.837688259081915, 0.0549683889839798, 0.892531563760713, 0.115890838671476, 0.400646223686635, 0.574344792868942, 0.0221788017079234, 0.0625251061283052, 0.708456309745088, 0.524330354528502, 0.28442710544914, 0.351325366878882, 0.852560516446829, 0.511335327289999, 0.328614609548822, 0.75226219650358, 0.137111845659092, 0.600668650586158, 0.57922185305506, 0.1421511862427, 0.144177082926035, 0.722248511854559, 0.459175070980564, 0.342041713418439, 0.795206875074655, 0.264847206417471, 0.436651576077566, 0.63165698456578, 0.210019383113831, 0.167321394896135, 0.754754480905831, 0.584055744577199, 0.0478321628179401, 0.557243207935244, 0.0593746453523636, 0.868778108386323, 0.13531103101559, 0.579364280216396, 0.651860991725698, 0.132745086913928, 0.305088681634516, 0.789050878025591, 0.705540773225948, 0.400778615381569, 0.110138579737395, 0.785793827846646, 0.682335862191394, 0.182795973028988, 0.349580513080582, 0.625476391986012, 0.258355360943824, 0.233044604072347, 0.554269529180601, 0.341256843181327, 0.068685329053551, 0.241441030288115, 0.613520823419094, 0.00406299694441259, 0.886562775354832, 0.166901897871867, 0.456736461492255, 0.733692167326808, 0.170693729538471, 0.4587719400879, 0.00300851417705417, 0.966303054941818, 0.604454652871937, 0.24935623863712, 0.578650478040799, 0.98110506683588, 0.546460314886644, 0.821501028956845, 0.707543936325237, 0.879164916230366, 0.527564749820158, 0.984203960048035, 0.982831743778661, 0.822351602371782, 0.909126653335989, 0.30845186393708, 0.107682062312961, 0.25311600859277, 0.552593050757423, 0.424661178840324, 0.0163380375597626, 0.330882519250736, 0.160275706788525, 0.359944624360651, 0.244194051483646, 0.385104280197993, 0.209439605474472, 0.0244849782902747, 0.872581242583692, 0.206714047351852, 0.548615642590448, 0.753801500191912, 0.348189548589289, 0.977848895126954, 0.505453103221953, 0.90857369126752, 0.127753159729764, 0.577086573000997, 0.725767329800874, 0.88620223174803, 0.428807441843674, 0.855906792683527, 0.078833722975105, 0.399178224615753, 0.760258640861139, 0.219960103509948, 0.737067948328331, 0.153905894840136, 0.412171149393544), actual = c(0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"