Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.5474795        0.6045284 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5262571        0.5262571        0.4888688 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.990875653456897, 0.62395591661334, 0.313924205023795, 0.9150706527289, 0.451947629684582, 0.221566211665049, 0.640814417274669, 0.683815728407353, 0.362540309783071, 0.821477115387097, 0.125721039250493, 0.0596762655768543, 0.110672807786614, 0.385886132949963, 0.761116805719212, 0.859500441234559, 0.168541489169002, 0.203857755521312, 0.643203544197604, 0.160465267021209, 0.762642082525417, 0.259046996943653, 0.515331509057432, 0.182045387802646, 0.669775464572012, 0.0718103412073106, 0.021767814643681, 0.0637376138474792, 0.611295339884236, 0.603686161106452, 0.36892752838321, 0.931134729180485, 0.271841251756996, 0.567607286851853, 0.802013086155057, 0.14295281143859, 0.581435505766422, 0.928999731549993, 0.441396980080754, 0.328844930743799, 0.754578202730045, 0.995860889088362, 0.580386608373374, 0.382643933873624, 0.492439542897046, 0.542043659603223, 0.369652030756697, 0.244983735028654, 0.109769575996324, 0.0550804547965527, 0.545217650942504, 0.689622870646417, 0.961248215055093, 0.98277346836403, 0.0870904519688338, 0.440194719936699, 0.135574511718005, 0.749771992443129, 0.373650255380198, 0.313250224338844, 0.405847461195663, 0.654357715044171, 0.742139196256176, 0.884695464977995, 0.349126775283366, 0.931463043903932, 0.0186185510829091, 0.639644023030996, 0.920378773706034, 0.573226581094787, 0.601876747095957, 0.801693965448067, 0.807312052231282, 0.578077249461785, 0.984988629352301, 0.265244498848915, 0.272376228123903, 0.311248244019225, 0.515163659118116, 0.529389279428869, 0.926463047973812, 0.722830145154148, 0.64297756110318, 0.525327914394438, 0.517680952558294, 0.435270329704508, 0.47139328578487, 0.27524702786468, 0.655935200164095, 0.181976684834808, 0.617727278033271, 0.229379212716594, 0.0761083192192018, 0.090407362440601, 0.327725081704557, 0.397235553013161, 0.720672610448673, 0.388280811486766, 0.581476423889399, 0.703893848694861, 0.492327195126563, 0.528312468668446, 0.705142692662776, 0.904570920625702, 0.216910083778203, 0.852610783651471, 0.450793343130499, 0.314856606535614, 0.0862152432091534, 0.997288913233206, 0.688861354254186, 0.844117421656847, 0.979619130725041, 0.784419662551954, 0.267498709261417, 0.776135088643059, 0.0528497395571321, 0.154673262266442, 0.0181042954791337, 0.43057697894983, 0.0550603508017957, 0.0275765885598958, 0.485576107399538, 0.880156479077414, 0.515157158719376, 0.432749133789912, 0.615938758244738, 0.6988626930397, 0.904119619401172, 0.0437787545379251, 0.210312343202531, 0.949155832873657, 0.448887523496523, 0.219609001884237, 0.699336290825158, 0.728944424772635, 0.0478291462641209, 0.977292788680643, 0.500874137040228, 0.324737127870321, 0.369911695364863, 0.479144544107839, 0.216324534965679, 0.167475429596379, 0.250274625839666, 0.963414739118889, 0.988303670892492, 0.716214934131131, 0.850536117097363, 0.847013490274549, 0.698272922541946, 0.20219182013534, 0.336163284024224, 0.891322142444551, 0.189441439462826, 0.110600336454809, 0.143299622926861, 0.82228898559697, 0.305619985796511, 0.97005659644492, 0.795800058171153, 0.508653632597998, 0.80802107299678, 0.479275679215789, 0.119725139345974, 0.502873813267797, 0.384510396048427, 0.689355307258666, 0.235825499985367, 0.644599029095843, 0.118409330025315, 0.233449852094054, 0.132213679375127, 0.990749922348186, 0.691498731030151, 0.154490676475689, 0.189111734274775, 0.469824134372175, 0.874779662350193, 0.849305187584832, 0.946188375120983, 0.750357144046575, 0.412325173383579, 0.298827434424311, 0.65924337040633, 0.49188348161988, 0.0472412565723062, 0.068985749501735, 0.14009089814499, 0.633944540284574, 0.889524313854054, 0.236001814017072, 0.13448926131241, 0.444269890198484, 0.388142921030521, 0.830722738057375, 0.200805489905179, 0.564046767540276, 0.0608732576947659, 0.235602129716426, 0.870920492568985, 0.155822388827801, 0.246421269606799, 0.655204250011593, 0.892301300540566, 0.89965962478891, 0.414960591821, 0.927199867554009, 0.149434781866148, 0.0992282242514193, 0.327425304800272, 0.553861222229898, 0.410576776834205, 0.221146438503638, 0.92166589666158, 0.992242840351537, 0.137680641375482, 0.85668235225603, 0.852583022089675, 0.0515853157266974, 0.763120286166668, 0.946635050699115, 0.856950172455981, 0.30880037532188, 0.299537291750312, 0.322167456848547, 0.324212106410414, 0.798090289114043, 0.0150962986517698, 0.0735920940060169, 0.319468984380364, 0.47186250728555, 0.102306919405237, 0.0727298946585506, 0.323035190114751, 0.497806375846267, 0.25718506006524, 0.683817530749366, 0.303195127751678, 0.352756329579279, 0.661537873791531, 0.537992640631273, 0.826089703012258, 0.352542503504083, 0.244514510966837, 0.0551500441506505, 0.0691317478194833, 0.762585800373927, 0.765113130910322, 0.434206092497334, 0.994023794308305, 0.957830143393949, 0.951081742532551, 0.535164447501302, 0.808372125262395, 0.576738244388252, 0.641980773536488, 0.277341569308192, 0.391890414291993, 0.765180013142526, 0.442976644728333, 0.908325964352116, 0.445196208078414, 0.16451383032836, 0.946157804690301, 0.598595902323723, 0.941493243444711, 0.461435962934047, 0.741047382820398, 0.448562605073676, 0.87116918922402, 0.41772896819748, 0.52804551157169, 0.678355887299404, 0.0580266322940588, 0.746699917828664, 0.582323114387691, 0.343997297110036, 0.997633698396385, 0.167601851979271, 0.315725712338462, 0.256595460232347, 0.0152358543127775, 0.213743134168908, 0.632770340424031, 0.61475060833618, 0.111537702847272, 0.865751922596246, 0.561843913514167, 0.436323772417381, 0.0226019897963852, 0.267373755807057, 0.261040651006624, 0.149948281701654, 0.817267572972924, 0.526036089519039, 0.989319714950398, 0.0888750597368926, 0.137543873628601, 0.197538628941402, 0.682347095105797, 0.114046807633713, 0.647825804306194, 0.571151364827529, 0.00365894939750433, 0.654527660692111, 0.864589117700234, 0.0937126975040883, 0.641078937100247, 0.793739630607888, 0.0739953080192208, 0.785500274039805, 0.0850065597333014, 0.95482710050419, 0.319917412707582, 0.865771361161023, 0.560975917382166, 0.39940443309024, 0.117159230168909, 0.615668952232227, 0.582507924409583, 0.973606330575421, 0.0109004981350154, 0.746224966598675, 0.434677381301299, 0.467172045726329, 0.584965691668913, 0.319577341200784, 0.54502892261371, 0.403147456469014, 0.192197874188423, 0.244532326236367, 0.872124900575727, 0.0454745190218091, 0.393715504091233, 0.0269385396968573, 0.113228022353724, 0.577322917524725, 0.162105270195752, 0.23797991219908, 0.105491657974198, 0.923947831382975, 0.289893766865134, 0.855708460323513, 0.124399616383016, 0.647413967875764, 0.77377056912519, 0.365035072667524, 0.0299447106663138, 0.597729621455073, 0.912354683270678, 0.244259590981528, 0.678940723417327, 0.691576377023011, 0.670583830447868, 0.661301208660007, 0.988484497880563, 0.389581188559532, 0.459437981015071, 0.523027488263324, 0.163524665171281, 0.972718121483922, 0.252271143719554, 0.0499740480445325, 0.418801684398204, 0.412196792429313, 0.0668566033709794, 0.528859285870567, 0.243929926771671, 0.805803747382015, 0.70821786718443, 0.30276475707069, 0.578999965684488, 0.911245885072276, 0.436277534812689, 0.310373662505299, 0.82664735801518, 0.0494919663760811, 0.501849159598351, 0.568922608392313, 0.246705277822912, 0.0824608036782593, 0.469345971709117, 0.551803176524118, 0.879239979665726, 0.250552453799173, 0.997937276028097, 0.894298833329231, 0.195486861281097, 0.44595678942278, 0.907000771258026, 0.569263857090846, 0.996063949540257, 0.313369295326993, 0.676894299685955, 0.83962082490325, 0.751856301212683, 0.628892296226695, 0.785275565693155, 0.844334149267524, 0.654252999462187, 0.33106053317897, 0.112867966992781, 0.516542680328712, 0.139930183067918, 0.134262496139854, 0.48036898067221, 0.650611054617912, 0.181956411106512, 0.0284812937024981, 0.567626697942615, 0.427723865024745, 0.663428513100371, 0.479347478598356, 0.653228340437636, 0.987561982357875, 0.745043316390365, 0.793304291786626, 0.521483345655724, 0.0483289984986186, 0.243338843574747, 0.671145385364071, 0.716140440432355, 0.530063469428569, 0.545434242812917, 0.222932177828625, 0.361959390575066, 0.427021962590516, 0.495031414786354, 0.232613515341654, 0.126106983050704, 0.643763174535707, 0.0716371650341898, 0.848650689702481, 0.329558062134311, 0.847974464064464, 0.278963589342311, 0.0530702078249305, 0.0479135622736067, 0.547611874295399, 0.175417479127645, 0.0383260904345661, 0.639975643018261, 0.172999700531363, 0.562186449300498, 0.523319938685745, 0.424106229795143, 0.475804088637233, 0.515345198567957, 0.667953809024766, 0.040844282368198, 0.358015663921833, 0.187024888116866, 0.0259687423240393, 0.152053916361183, 0.96373941982165, 0.122416506987065, 0.490198074840009, 0.898364015389234, 0.556215432239696, 0.488033261615783, 0.785674225538969, 0.38620101637207, 0.184554925886914, 0.350062682526186, 0.695696732262149, 0.611895297421142, 0.513754572952166, 0.79953270428814, 0.922465837327763, 0.666513456963003, 0.202268910128623, 0.789760413579643, 0.843671733280644, 0.543772314209491, 0.636625181417912, 0.93634009314701, 0.591152486624196, 0.764506181236356, 0.255420434987172, 0.50015998701565, 0.786718256538734, 0.0595764419995248, 0.204052958171815, 0.0672401713673025, 0.962060039397329, 0.467757833190262, 0.729790828423575, 0.0274705395568162, 0.934649756643921, 0.0931757087819278, 0.260270624188706, 0.59570742235519, 0.127578113228083, 0.844534339616075, 0.375812879297882, 0.26018940214999, 0.988417596323416, 0.952243493869901, 0.757898568408564, 0.293041804572567, 0.775385031709448, 0.330178760457784, 0.325199546292424, 0.270707890158519, 0.469829498324543, 0.219399630324915, 0.376933645224199, 0.3172535833437, 0.832208627834916, 0.757198948180303, 0.11200970155187, 0.111152541125193, 0.0076720027718693, 0.673249212792143, 0.684238344198093, 0.014391700271517, 0.537423694273457, 0.0965936325956136, 0.613268287386745, 0.0981657132506371, 0.816830767784268, 0.361612633103505, 0.376813804963604, 0.791566885076463, 0.0595836008433253, 0.886239959159866, 0.782050080597401, 0.347390888025984, 0.165730130858719, 0.0874484297819436, 0.646135126939043, 0.159127131802961, 0.13296449999325, 0.165395032148808, 0.26240160712041, 0.786306285299361, 0.14375138329342, 0.123822969151661, 0.504239534493536, 0.975380746647716, 0.690382270608097, 0.141949487617239, 0.157006014604121, 0.890135052381083, 0.211240350501612, 0.572472930653021, 0.122124290792271, 0.189264544053003, 0.918023675912991, 0.399342809803784, 0.70604315912351, 0.193195146974176, 0.100073753623292, 0.475387640297413, 0.350543844280764, 0.479530113050714, 0.844698084285483, 0.113528873771429, 0.849007129203528, 0.76355913747102, 0.0252667739987373, 0.898868302349001, 0.920393957290798, 0.84155950229615, 0.754191427258775, 0.799832981545478, 0.101237190887332, 0.636416300432757, 0.366299968212843, 0.757119045825675, 0.965202352032065, 0.257052726345137, 0.216946881730109, 0.434659909456968, 0.331865119049326, 0.763885802822188, 0.0742208706215024, 0.684584851376712, 0.535739437444136, 0.848543220665306, 0.278314156224951, 0.607827077852562, 0.480427916860208, 0.53888798924163, 0.462411686312407, 0.824160836404189, 0.0197127209976315, 0.820235839812085, 0.0644136108458042, 0.667402243241668, 0.623588257469237, 0.948985443916172, 0.0264297488611192, 0.732737688580528, 0.215797680662945, 0.235965189523995, 0.201783900149167, 0.0941595383919775, 0.812826022738591, 0.988950995961204, 0.883170336019248, 0.885322391754016, 0.814308769302443, 0.81895655230619, 0.921282252063975, 0.487894164631143, 0.4886781538371, 0.0691133260261267, 0.954950823215768, 0.246014645090327, 0.479605952044949, 0.32900236081332, 0.130580140976235, 0.699495882727206, 0.571970667922869, 0.885326254181564, 0.697178259026259, 0.340749573195353, 0.471571441739798, 0.233114420669153, 0.896173116052523, 0.891223081387579, 0.527767632855102, 0.0249027544632554, 0.391242194920778, 0.879943788284436, 0.807142998091877, 0.292649229522794, 0.488590152235702, 0.637606108561158, 0.516430214745924, 0.19563804497011, 0.594429661054164, 0.347823246847838, 0.632279714103788, 0.940636835992336, 0.595305200433359, 0.192508602282032, 0.78087335312739, 0.301244592526928, 0.357390721794218, 0.234831115463749, 0.839849879499525, 0.848541141021997, 0.160316398134455, 0.217307747574523, 0.487944322172552, 0.647164318244904, 0.45009701512754, 0.183362011564896, 0.657387635903433, 0.485948191722855, 0.451629609102383, 0.0642192966770381, 0.473360870964825, 0.178084781626239, 0.103034381987527, 0.822947667678818, 0.377387573942542, 0.382894236827269, 0.571266381070018, 0.183959410525858, 0.193064762046561, 0.96372016472742, 0.940639952430502, 0.52570301736705, 0.282622165279463, 0.847634271485731, 0.732045952929184, 0.544783073710278, 0.189018184086308, 0.939605407649651, 0.579656299902126, 0.933248461922631, 0.307616403093562, 0.392491366248578, 0.0422830767929554, 0.375747913727537, 0.948918478563428, 0.4179704899434, 0.819897313835099, 0.723067393526435, 0.180037947604433, 0.590943953953683, 0.577539626508951, 0.0312693323940039, 0.187710303580388, 0.413458846043795, 0.555432904278859, 0.348118490539491, 0.78559972718358, 0.73348059784621, 0.0457837614230812, 0.533414676086977, 0.220683834981173, 0.666745287831873, 0.169982525985688, 0.672113155247644, 0.49026643880643, 0.808514004806057, 0.361007306259125, 0.641835562651977, 0.884474499616772, 0.764921963447705, 0.377752436790615, 0.950153331272304, 0.497998756123707, 0.473819470498711, 0.89921347098425, 0.0838641345035285, 0.882924874313176, 0.935409788275138, 0.434104705927894, 0.59919983940199, 0.607819435186684, 0.767434736480936, 0.922783614136279, 0.0169949766714126, 0.861103693954647, 0.917962099425495, 0.338129905285314, 0.820946872234344, 0.0547878581564873, 0.364212116692215, 0.117353826761246, 0.43704771483317, 0.0994016602635384, 0.499241806566715, 0.780692025087774, 0.0579410996288061, 0.0490267528221011, 0.529291338287294, 0.972222496522591, 0.547844325192273, 0.943392870947719, 0.642811662983149, 0.0535037964582443, 0.160908210556954, 0.155386334983632, 0.787476145895198, 0.106075995368883, 0.617884787963703, 0.00599874998442829, 0.936087219510227, 0.232269844273105, 0.310418702661991, 0.83137216209434, 0.256139550358057, 0.554484093328938, 0.810996717307717, 0.7369342839811, 0.139130290364847, 0.451677701668814, 0.18200183683075, 0.673152758972719, 0.0860623265616596, 0.936480681644753, 0.858610252849758, 0.303302362095565, 0.834155551390722, 0.0251092864200473, 0.781582708004862, 0.850729236146435, 0.709018499823287, 0.992030798923224, 0.701820096001029, 0.174136387417093, 0.528760582208633, 0.372220832156017, 0.827044593170285, 0.479313029907644, 0.063782159704715, 0.688046254916117, 0.366550930775702, 0.734245955944061, 0.239899475127459, 0.862145945662633, 0.336591612081975, 0.525863037444651, 0.677493955707178, 0.242186991032213, 0.565330745186657, 0.891630270052701, 0.512514715082943, 0.561743875266984, 0.571124121313915, 0.484024042030796, 0.53939760569483, 0.779269265476614, 0.302306007826701, 0.827373610110953, 0.9417169096414, 0.759694809792563, 0.965041219256818, 0.187345501733944, 0.00352147128432989, 0.579012430971488, 0.495359146501869, 0.542431704932824, 0.367460899520665, 0.28805563505739, 0.477450705366209, 0.170360461110249, 0.62101667211391, 0.666370264720172, 0.284004620509222, 0.825272803893313, 0.663324416382238, 0.0993811336811632, 0.375365974614397, 0.873419194016606, 0.340354222804308, 0.31481542927213, 0.530436626868322, 0.786462978692725, 0.829616670496762, 0.795483228983358, 0.292972009861842, 0.637784794904292, 0.208652753150091, 0.898668082430959, 0.474602952133864, 0.312549929134548, 0.841583991190419, 0.729830150492489, 0.532956225099042, 0.417463858611882, 0.00568085932172835, 0.23883763817139, 0.780037327902392, 0.755961341550574, 0.323042410425842, 0.959943915950134, 0.959073320263997, 0.981859952444211, 0.63225580798462, 0.593938631704077, 0.518891144543886, 0.605064219562337, 0.937235568650067, 0.744478475069627, 0.546798232011497, 0.160128966206685, 0.374513390939683, 0.520403967704624, 0.292230965802446, 0.396782488096505, 0.827236569719389, 0.620693433564156, 0.308714375132695, 0.878370008897036, 0.513166137970984, 0.36623493116349, 0.34442678373307, 0.330076878890395, 0.355367883574218, 0.622306927572936, 0.983245536917821, 0.969153606100008, 0.256412576185539, 0.401402228744701, 0.362602622248232, 0.715625992743298, 0.581743025220931, 0.0506266129668802, 0.179729494964704, 0.75646524480544, 0.240456220461056, 0.292868240969256, 0.357160358456895, 0.935363578144461, 0.97721166908741, 0.296055653132498, 0.837523390771821, 0.886044223792851, 0.2643379624933, 0.996347985230386, 0.536648923298344, 0.76333896862343, 0.798784275306389, 0.535104630980641, 0.0247559603303671, 0.689125027041882, 0.492391894571483, 0.284734627697617, 0.639856504043564, 0.389045945834368, 0.0270003550685942, 0.33314002933912, 0.18397101922892, 0.127723209792748, 0.868288531899452, 0.965425389586017, 0.963065828662366, 0.483421857701614, 0.597930769668892, 0.827554915333167, 0.156642107758671, 0.481078607728705, 0.657353922724724, 0.323510760907084, 0.452543714782223, 0.295209166128188, 0.479897206183523, 0.576054609147832, 0.360516900662333, 0.290120959514752, 0.463504387997091, 0.607151300879195, 0.149106411729008, 0.632750962628052, 0.760019295383245, 0.552388103911653, 0.945745344273746, 0.396467766724527, 0.420931615866721, 0.528087598271668, 0.730985823785886, 0.130159923573956, 0.10949162975885, 0.683299653232098, 0.335930048022419, 0.737844389863312, 0.500718607567251, 0.02164490474388, 0.952054297551513, 0.813815418863669, 0.501607002457604, 0.770982103422284, 0.496378497686237, 0.94947526534088, 0.202643443830311, 0.246033335337415, 0.108917734818533, 0.747489128727466, 0.797219868283719, 0.686197204049677, 0.769878817023709, 0.0957387718372047, 0.46294431318529, 0.936246930854395, 0.447181537747383, 0.940136829623953, 0.94242531969212, 0.644882762571797, 0.62489398336038, 0.521266044583172, 0.36859921249561, 0.522335022222251, 0.334101356798783, 0.0535310194827616, 0.90509879286401, 0.442737491801381, 0.802116732578725, 0.945700295036659, 0.838070422178134, 0.951337149366736, 0.481208104407415, 0.399805617518723, 0.489252995932475, 0.65706124342978, 0.219307226827368, 0.480632779421285, 0.289135997649282, 0.203288369579241, 0.0943301238585263, 0.254460077499971, 0.196827929699793, 0.910557001363486, 0.51899846107699, 0.668609102256596, 0.767541635315865, 0.688571638427675, 0.0757911871187389, 0.974336730781943, 0.673506932333112, 0.416070821229368, 0.653650472173467, 0.0632872772403061, 0.593876356491819, 0.704911065520719, 0.805601462721825, 0.375956896925345, 0.42079058312811, 0.246904087485746, 0.0635942472144961, 0.036263833520934, 0.345300724729896, 0.509226874448359, 0.0950842259917408, 0.120483271311969, 0.711974281817675, 0.444169427733868, 0.531659664353356, 0.404719472164288, 0.240874145412818, 0.192007313249633, 0.501816652482376, 0.941149555612355), actual = c(1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.274680254049599, 0.000189711339771748, 0.264597087632865, 0.714260558597744, 0.510816210182384, 0.855846581049263, 0.733159748837352, 0.657520139589906, 0.23089635069482, 0.764047552132979, 0.900826022727415, 0.00129692791961133, 0.796104966895655, 0.826469516847283, 0.922140494920313, 0.200620034011081, 0.759650174062699, 0.609063217416406, 0.939911350840703, 0.654089466203004, 0.574552217964083, 0.725563434883952, 0.407635391689837, 0.187886287458241, 0.183611160609871, 0.219728755299002, 0.422586873173714, 0.83112114132382, 0.857135884696618, 0.846323828212917, 0.976685284404084, 0.382078300230205, 0.903949971310794, 0.0378471147269011, 0.027933839475736, 0.283136991318315, 0.295834339922294, 0.111714636674151, 0.953759845579043, 0.819708951748908, 0.290948542766273, 0.705719054443762, 0.355436270823702, 0.847643950721249, 0.553392396308482, 0.401849320158362, 0.604156678076833, 0.58657376235351, 0.250951187452301, 0.994992390740663, 0.468930040486157, 0.160064374096692, 0.663437879877165, 0.765511033823714, 0.231387888779864, 0.841178147122264, 0.544331294018775, 0.911853455938399, 0.0959352271165699, 0.732438367558643, 0.571644903393462, 0.350928144529462, 0.467702010180801, 0.0295365350320935, 0.848113242071122, 0.420433009741828, 0.607937239576131, 0.41746720764786, 0.223217584658414, 0.220334175275639, 0.197377726901323, 0.450289664324373, 0.988055953988805, 0.937577226664871, 0.476998905651271, 0.135197964729741, 0.358822502428666, 0.260216865222901, 0.109861118951812, 0.580699619837105, 0.0247982263099402, 0.838422993430868, 0.389095905702561, 0.577987446682528, 0.0628430705983192, 0.135698911035433, 0.805203928379342, 0.127327059861273, 0.3762468455825, 0.871393579989672, 0.837466971948743, 0.481192029779777, 0.915185386547819, 0.634097139118239, 0.5909637324512, 0.308874499518424, 0.699708829401061, 0.476108203176409, 0.589568190975115, 0.443226055707783, 0.0278769151773304, 0.173519145697355, 0.699959642020985, 0.682540117762983, 0.228945335373282, 0.202082552248612, 0.130908615654334, 0.749928303761408, 0.13336222502403, 0.0585387500468642, 0.73980092536658, 0.417337730061263, 0.556317075621337, 0.0079158169683069, 0.73998241382651, 0.167582127964124, 0.482978692743927, 0.111579374177381, 0.799541560467333, 0.977585002314299, 0.923880627844483, 0.554458783240989, 0.328363810898736, 0.688842908479273, 0.0692803524434566, 0.493396516190842, 0.784174099564552, 0.831990190315992, 0.338648732285947, 0.230969172669575, 0.291528546484187, 0.662819272605702, 0.243926265509799, 0.467208117246628, 0.993497144198045, 0.537209497299045, 0.338797812117264, 0.709678616607562, 0.383299698121846, 0.0949207891244441, 0.7960235606879, 0.392255885526538, 0.510469327447936, 0.926159440074116, 0.246568836504593, 0.173361178254709, 0.239322407403961, 0.922642703400925, 0.420671519823372, 0.634964470984414, 0.618482588790357, 0.0959900594316423, 0.413612354313955, 0.0536145165096968, 0.670758366817608, 0.978777105920017, 0.288734486792237, 0.356351320398971, 0.727805119706318, 0.0481447835918516, 0.0988314033020288, 0.913953189505264, 0.890670585911721, 0.427700065076351, 0.206177934771404, 0.81670138426125, 0.499622701201588, 0.902084796223789, 0.902452065842226, 0.40260644396767, 0.140247236005962, 0.805269327480346, 0.702558536315337, 0.438838626956567, 0.707101846812293, 0.709354810416698, 0.0182161801494658, 0.702707157004625, 0.302248690277338, 0.172384298406541, 0.531513270456344, 0.798255518078804, 0.571376331383362, 0.130251072579995, 0.214722544420511, 0.658850677777082, 0.301781416172162, 0.679821175290272, 0.253749544965103, 0.429343458963558, 0.522695289924741, 0.190721659455448, 0.953896148595959, 0.00554824690334499, 0.726436495315284, 0.425880239577964, 0.00395870883949101, 0.850558680715039, 0.920500656822696, 0.984611444640905, 0.75144461216405, 0.506063519977033, 0.349437317112461, 0.401088437531143, 0.762341414345428, 0.751314179738984, 0.193908066023141, 0.306887804530561, 0.860860649030656, 0.758629330433905, 0.12857926543802, 0.663862260524184, 0.314810923300683, 0.955960871186107, 0.915939236059785, 0.920729690697044, 0.00755097973160446, 0.00305384653620422, 0.434149112785235, 0.0648744131904095, 0.878021012060344, 0.866404608124867, 0.0358545444905758, 0.311752205481753, 0.463407753733918, 0.0122212595306337, 0.361660023219883, 0.645934283267707, 0.909419142641127, 0.464318688027561, 0.385758436052129, 0.893595958594233, 0.449783229734749, 0.64625638932921, 0.326755389804021, 0.274706943426281, 0.0231809420511127, 0.586738847428933, 0.900438879383728, 0.0592203494161367, 0.121872374787927, 0.246306004235521, 0.483029534108937, 0.300676992628723, 0.293346881400794, 0.485917655052617, 0.00743948994204402, 0.244653527624905, 0.927055662497878, 0.0782802267931402, 0.207046198891476, 0.870643995003775, 0.077177828643471, 0.0120417228899896, 0.0793782791588455, 0.155902384081855, 0.0591435637325048, 0.806379181565717, 0.131813449552283, 0.261351245921105, 0.103921222267672, 0.955497928429395, 0.558104296214879, 0.529499670956284, 0.295445877127349, 0.292025697184727, 0.793682751478627, 0.333304864587262, 0.135370939970016, 0.721610905602574, 0.64945432706736, 0.421700815437362, 0.312310605775565, 0.508401127764955, 0.90918378601782, 0.693741413066164, 0.619539235718548, 0.77360764099285, 0.746425116201863, 0.0955738215707242, 0.672039210796356, 0.0984765198081732, 0.102526057977229, 0.225549574010074, 0.0334231199230999, 0.334171804832295, 0.649010291090235, 0.80118433595635, 0.1786852509249, 0.227355272974819, 0.767341912724078, 0.780583201441914, 0.440608833683655, 0.0707621886394918, 0.273474491201341, 0.527883043978363, 0.0421860185451806, 0.0359354824759066, 0.990417553577572, 0.0227020855527371, 0.0790144398342818, 0.982757591642439, 0.902615809347481, 0.034690456930548, 0.526795512763783, 0.011286482680589, 0.45390689256601, 0.0496090590022504, 0.761732107959688, 0.593445285921916, 0.506152209127322, 0.747104396810755, 0.3870859015733, 0.833846797700971, 0.381097238743678, 0.944306710036471, 0.291847400832921, 0.986900860443711, 0.259049563901499, 0.921286179218441, 0.365454369224608, 0.990422274917364, 0.908076124731451, 0.169382060412318, 0.42203181819059, 0.0417077916208655, 0.509615318384022, 0.204087594756857, 0.0379272671416402, 0.492937701288611, 0.503662484930828, 0.722249823855236, 0.403196508064866, 0.818692188244313, 0.394881451735273, 0.0760413641110063, 0.0378657686524093, 0.136825118213892, 0.093423143029213, 0.966848359443247, 0.564906060229987, 0.876333107473329, 0.117374296300113, 0.329970824997872, 0.871274837292731, 0.999307058751583, 0.752612021286041, 0.947545072296634, 0.348631927976385, 0.164065790129825, 0.502657607663423, 0.953984417021275, 0.128553934395313, 0.351130333496258, 0.276046374579892, 0.0890667866915464, 0.546792862238362, 0.770149803487584, 0.649719418259338, 0.900381304323673, 0.119929110398516, 0.100313918432221, 0.667019543237984, 0.562375161563978, 0.478823923971504, 0.293251767288893, 0.660310537088662, 0.550113512435928, 0.584975899895653, 0.170388058526441, 0.425767828244716, 0.595288353506476, 0.50138274836354, 0.60308882733807, 0.294297972228378, 0.900557111715898, 0.851776086958125, 0.700317438459024, 0.780450040008873, 0.422848072368652, 0.257714780978858, 0.968747155042365, 0.842583208112046, 0.0794144512619823, 0.663644434884191, 0.186874229693785, 0.286232899641618, 0.203817720757797, 0.872989323688671, 0.72585975122638, 0.630047839600593, 0.468602599110454, 0.184957213466987, 0.30499715055339, 0.475497808074579, 0.663463375531137, 0.0113274196628481, 0.870391130913049, 0.383919093292207, 0.765276711899787, 0.73766230721958, 0.121120177675039, 0.793530226685107, 0.913262419169769, 0.44351349840872, 0.394405010156333, 0.475748020689934, 0.808808503672481, 0.98503631609492, 0.510266793193296, 0.114943318301812, 0.446534529561177, 0.148591371718794, 0.595693272771314, 0.748475006548688, 0.25541387591511, 0.948184994747862, 0.16359503287822, 0.248399669537321, 0.577840691898018, 0.362351644551381, 0.54230406251736, 0.651959928683937, 0.162928046192974, 0.87085232976824, 0.628653227584437, 0.475024533458054, 0.0454167441930622, 0.965807208325714, 0.781142274849117, 0.0874568971339613, 0.463562501128763, 0.588741685962304, 0.639059252338484, 0.0167769212275743, 0.0849123422522098, 0.473366708727553, 0.742492005461827, 0.564586433582008, 0.347565436735749, 0.698464013403282, 0.665600912412629, 0.696968115400523, 0.638821252854541, 0.0862635322846472, 0.31885186699219, 0.848236873978749, 0.523594522383064, 0.906056070933118, 0.424590340582654, 0.44758107769303, 0.785066477721557, 0.472808229504153, 0.641544915502891, 0.848048838088289, 0.722222079057246, 0.563554686494172, 0.21717638685368, 0.876530901761726, 0.664829320972785, 0.819734379649162, 0.470919264946133, 0.419434584211558, 0.398215459659696, 0.500256427330896, 0.489343746798113, 0.337525764480233, 0.738385063130409, 0.389716926962137, 0.296360195148736, 0.889564465731382, 0.086641764966771, 0.685968475649133, 0.558310435153544, 0.566143175121397, 0.0636815913021564, 0.270376757718623, 0.714890983654186, 0.558958731824532, 0.81384751922451, 0.0696604845579714, 0.607478693593293, 0.365341385360807, 0.787306134123355, 0.254025995964184, 0.045343364123255, 0.0238485331647098, 0.623777294764295, 0.965889750048518, 0.494110811501741, 0.00508897076360881, 0.397909930674359, 0.303130198735744, 0.414282756624743, 0.0552533387672156, 0.681593298679218, 0.344985173083842, 0.901020622812212, 0.318414688576013, 0.705866860924289, 0.210313239134848, 0.598211898934096, 0.213286010082811, 0.248947302578017, 0.0572151325177401, 0.568396482849494, 0.118958755163476, 0.511638652766123, 0.580798670416698, 0.699852925492451, 0.616482648300007, 0.664575687609613, 0.00946485926397145, 0.569398826919496, 0.154970466392115, 0.0491254094522446, 0.692574960179627, 0.877068574540317, 0.568098654272035, 0.190974344266579, 0.766292395535856, 0.121052931295708, 0.0565623696893454, 0.411219671135768, 0.571831145789474, 0.678634729702026, 0.655041565420106, 0.504786500940099, 0.466701824450865, 0.623963119229302, 0.306507264031097, 0.583414879860356, 0.912673109676689, 0.702028379542753, 0.909163062460721, 0.454607570543885, 0.823622772237286, 0.473021493060514, 0.90412810514681, 0.801911982242018, 0.749728061724454, 0.491577301174402, 0.911620169179514, 0.982730248244479, 0.696394255617633, 0.574503207113594, 0.361383204115555, 0.479853961151093, 0.183330731699243, 0.557992942864075, 0.121822151122615, 0.352822834858671, 0.120969805167988, 0.271856681676582, 0.664276967989281, 0.233442790107802, 0.417356317630038, 0.986243241699412, 0.328906680690125, 0.9946246703621, 0.0243353703990579, 0.590298260096461, 0.675956287188455, 0.435954513493925, 0.506066876230761, 0.789632185362279, 0.145075993379578, 0.085804586764425, 0.107989085838199, 0.566377449780703, 0.614454074762762, 0.655400847783312, 0.135086053516716, 0.898125907871872, 0.664689667755738, 0.688691948074847, 0.450634893961251, 0.559503971366212, 0.756606127601117, 0.159384890226647, 0.736794530414045, 0.394487789366394, 0.017198849003762, 0.184407072607428, 0.761334562674165, 0.708406252786517, 0.877858178922907, 0.407814555102959, 0.912038460141048, 0.0124225791078061, 0.160178095567971, 0.328836251050234, 0.42405970278196, 0.40318907238543, 0.58671517809853, 0.606249202508479, 0.0248953127302229, 0.252038558246568, 0.687841784441844, 0.978740205522627, 0.898760999320075, 0.0419591637328267, 0.293725090799853, 0.924421541858464, 0.604417022317648, 0.948487911373377, 0.617616177769378, 0.030134353088215, 0.558812467847019, 0.488996546948329, 0.183303650235757, 0.347135765943676, 0.63015774032101, 0.0457163790706545, 0.799873436801136, 0.507332114968449, 0.083636398660019, 0.412079556146637, 0.739367625210434, 0.682164827128872, 0.793068514438346, 0.733788411132991, 0.930121590616181, 0.114397049648687, 0.39146693632938, 0.846114527899772, 0.0818467719946057, 0.849645562935621, 0.183376403292641, 0.516820273594931, 0.738045120146126, 0.955393155803904, 0.555862097302452, 0.42843031254597, 0.656029846286401, 0.752540867542848, 0.548894674284384, 0.417630845913664, 0.986395881511271, 0.556879112962633, 0.450665418989956, 0.957526685437188, 0.879861867986619, 0.165370030794293, 0.475726661505178, 0.327422900125384, 0.00207329750992358, 0.473601929144934, 0.16451531345956, 0.816584323067218, 0.0406223142053932, 0.381059657549486, 0.226079999702051, 0.871409364510328, 0.774402634706348, 0.788400353631005, 0.768071985105053, 0.395197447389364, 0.750791647704318, 0.432802858063951, 0.788598782150075, 0.801204882794991, 0.0452150027267635, 0.289020730182528, 0.659619280602783, 0.696959356078878, 0.31338367331773, 0.433228090638295, 0.148350288858637, 0.242477958556265, 0.0378478888887912, 0.174230600940064, 0.967637381982058, 0.349383642431349, 0.785988419316709, 0.532060721190646, 0.204923521261662, 0.66037891083397, 0.294788502855226, 0.100555439479649, 0.759139886591583, 0.85484572686255, 0.126785523723811, 0.252971044974402, 0.790701257996261, 0.983218303881586, 0.494492896832526, 0.991871856385842, 0.379621405852959, 0.333604489220306, 0.573599521303549, 0.464275875128806, 0.64760569203645, 0.15802470385097, 0.0995610055979341, 0.686182663077489, 0.557459014235064, 0.0013996260240674, 0.41037572780624, 0.214943516068161, 0.582097300095484, 0.371798662701622, 0.376194581156597, 0.809460927965119, 0.954797186190262, 0.29598798789084, 0.225570488488302, 0.497986835660413, 0.916265497216955, 0.42709941836074, 0.667462653946131, 0.238953593419865, 0.414321872172877, 0.77756590093486, 0.789927925681695, 0.41616179025732, 0.469040622003376, 0.963034049374983, 0.617778783431277, 0.501491672592238, 0.994821161730215, 0.570073672337458, 0.0374575855676085, 0.859630253864452, 0.329815400764346, 0.118636570405215, 0.400625773705542, 0.620917458785698, 0.725087627302855, 0.871378101874143, 0.294501243857667, 0.935277473414317, 0.267964670434594, 0.71003879699856, 0.345030404161662, 0.102194192819297, 0.134527788264677, 0.456210673553869, 0.734991851262748, 0.220425571547821, 0.942792147630826, 0.66323269973509, 0.427642134483904, 0.421561740338802, 0.444530333159491, 0.4772297157906, 0.724633224774152, 0.983321704203263, 0.787112952210009, 0.868344286922365, 0.46343968831934, 0.643658840795979, 0.468032649718225, 0.767404007492587, 0.17906215111725, 0.597549059428275, 0.0432979369070381, 0.463073773542419, 0.585165443364531, 0.161319509614259, 0.746865238063037, 0.831218550214544, 0.541072627995163, 0.0947983202058822, 0.181510657770559, 0.24453697912395, 0.923169335816056, 0.227494637481868, 0.228999732527882, 0.202238362049684, 0.551712646847591, 0.707270223181695, 0.389632276957855, 0.406555178342387, 0.598886252846569, 0.361289366846904, 0.649338828166947, 0.794796759961173, 0.826415178133175, 0.721109883394092, 0.254178213188425, 0.567989923991263, 0.180653311079368, 0.00799753842875361, 0.446210282389075, 0.442946737166494, 0.59316966868937, 0.111295659793541, 0.127608368173242, 0.182327315676957, 0.610458663897589, 0.610060100443661, 0.87829451286234, 0.796551918145269, 0.421878924127668, 0.326466658618301, 0.0259725865907967, 0.602341070538387, 0.0876332365442067, 0.704584117280319, 0.0251098214648664, 0.454053195426241, 0.0525189691688865, 0.0376974325627089, 0.627954344730824, 0.437803929438815, 0.365745690185577, 0.165896211285144, 0.104091704124585, 0.396092545008287, 0.777823423966765, 0.102017430355772, 0.849925253307447, 0.233075562864542, 0.893740500323474, 0.715346290497109, 0.881748190848157, 0.408891834551468, 0.443772447761148, 0.76202958682552, 0.162163980538025, 0.35333009227179, 0.0332790350075811, 0.902160473866388, 0.766572830267251, 0.627188955666497, 0.354924379382282, 0.818961496232077, 0.0382860172539949, 0.699740125332028, 0.175454578595236, 0.499337859451771, 0.938829149119556, 0.615270402515307, 0.68797700596042, 0.640532581834123, 0.97769898083061, 0.269847507122904, 0.735925633693114, 0.902763313613832, 0.190496597671881, 0.689217395382002, 0.793311331653967, 0.153060567099601, 0.431328813545406, 0.457677443744615, 0.278590820729733, 0.895268060499802, 0.830706269945949, 0.523150121793151, 0.015525293070823, 0.586676894221455, 0.625450862338766, 0.0365995201282203, 0.58223522733897, 0.114055030047894, 0.519225306343287, 0.000825914088636637, 0.94891214882955, 0.0481620030477643, 0.914814396994188, 0.160091414349154, 0.427588638383895, 0.903124471427873, 0.803602550411597, 0.816294701071456, 0.692865276010707, 0.405989446677268, 0.169982958817855, 0.277374869212508, 0.595942324725911, 0.850644994759932, 0.518682097317651, 0.207379805622622, 0.298128157155588, 0.434217721689492, 0.0892216889187694, 0.0838278131559491, 0.584784964332357, 0.214049813803285, 0.96034020325169, 0.794894246850163, 0.4782766897697, 0.737364217173308, 0.157633835216984, 0.498145191231743, 0.486862377962098, 0.187717257766053, 0.367265846114606, 0.0366908137220889, 0.294849388301373, 0.178352528717369, 0.738026595441625, 0.181522579398006, 0.896442029392347, 0.636044225655496, 0.330689085647464, 0.932227866956964, 0.254878056002781, 0.0454702104907483, 0.989719296339899, 0.261448821751401, 0.977861697087064, 0.277792169712484, 0.775408433750272, 0.345252454513684, 0.438119124621153, 0.502729892963544, 0.872840730007738, 0.457074458710849, 0.945849961368367, 0.722558591514826, 0.0107445300091058, 0.480783224338666, 0.796353810001165, 0.0600056699477136, 0.00164136663079262, 0.202498292084783, 0.662586270831525, 0.334463618928567, 0.281747492263094, 0.893138909013942, 0.267083117971197, 0.485104323131964, 0.115805916953832, 0.194410940865055, 0.0629940768703818, 0.735766329336911, 0.416774161858484, 0.307538016000763, 0.850028431043029, 0.384753883350641, 0.912676072446629, 0.429113677702844, 0.211000988259912, 0.493485398357734, 0.0711533685680479, 0.76097470219247, 0.695332982344553, 0.49666347540915, 0.0560945810284466, 0.619871986797079, 0.133977824822068, 0.252818089909852, 0.18751553632319, 0.0777643357869238, 0.544266821583733, 0.120117600075901, 0.497727887937799, 0.472016138257459, 0.816346919862553, 0.621668230509385, 0.412610324798152, 0.948904232354835, 0.605655841063708, 0.447875120211393, 0.180143746780232, 0.596739965723827, 0.777316093444824, 0.0395754666533321, 0.294887351570651, 0.462688821135089, 0.997831778600812, 0.53770871902816, 0.458160132868215, 0.939412768464535, 0.0247424568515271, 0.480781339341775, 0.527838794980198, 0.973596513736993, 0.673869597958401, 0.90884691500105, 0.917390430811793, 0.0636715677101165, 0.678919995902106, 0.289428570540622, 0.964228481287137, 0.810020504519343, 0.359253223286942, 0.438600016292185, 0.0448184257838875, 0.290359110105783, 0.580561588285491, 0.158968173665926, 0.802000468131155, 0.226737532764673, 0.843203728552908, 0.673909867415205, 0.185881615383551, 0.203635087935254, 0.0357318955939263, 0.568112609675154, 0.876223657513037, 0.797208172501996, 0.825853560119867, 0.711880254559219, 0.0182283062022179, 0.542349186958745, 0.0149525401648134, 0.386649011867121), actual = c(0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"