Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4435485        0.4711864 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5027344        0.5027344        0.5283784 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.495805381098762, 0.493457053322345, 0.714861248619854, 0.327362421900034, 0.214954760856926, 0.840679886983708, 0.646016970276833, 0.21824010508135, 0.739026219351217, 0.712368186330423, 0.714228031923994, 0.184035054640844, 0.249547597719356, 0.535660393536091, 0.0966188155580312, 0.610166002064943, 0.438560921000317, 0.558273744536564, 0.803615238517523, 0.329566982807592, 0.793647244106978, 0.317079852800816, 0.143687525531277, 0.606612999225035, 0.170994397019967, 0.0981850305106491, 0.396202966570854, 0.715309423161671, 0.682571003912017, 0.572541850153357, 0.953984020045027, 0.51136510935612, 0.30213479930535, 0.14890692429617, 0.492534564109519, 0.92889224877581, 0.995479059172794, 0.818999191746116, 0.685197344981134, 0.656958792824298, 0.796267675934359, 0.751458629965782, 0.36758276168257, 0.33627662807703, 0.643555149435997, 0.00141292857006192, 0.0879266760312021, 0.645220617763698, 0.181295865913853, 0.127753828419372, 0.766503271413967, 0.0165968781802803, 0.0473357224836946, 0.507761224638671, 0.641901188762859, 0.952243912033737, 0.82831283705309, 0.943395301233977, 0.994127693586051, 0.35275395703502, 0.374431711854413, 0.992142707575113, 0.817703702487051, 0.199825446587056, 0.773885434027761, 0.619024760089815, 0.374228307511657, 0.232761630089954, 0.428276529302821, 0.586109845666215, 0.854169623926282, 0.197412207257003, 0.676001891028136, 0.64027155819349, 0.166389962425455, 0.0902116196230054, 0.654938273830339, 0.535849877167493, 0.257896006805822, 0.783971362980083, 0.835084680235013, 0.358827740652487, 0.136618209071457, 0.28549376106821, 0.886357555398718, 0.997916342923418, 0.902600316330791, 0.242890045046806, 0.821916236076504, 0.251193954143673, 0.361673040548339, 0.677718677092344, 0.542280720779672, 0.0487872147932649, 0.441145285498351, 0.377913928357884, 0.831050128210336, 0.50991359911859, 0.304623000323772, 0.632095545995981, 0.780951894121245, 0.826808907557279, 0.999272990971804, 0.401875517563894, 0.622630309546366, 0.0433276745025069, 0.3893657468725, 0.355014670873061, 0.246832965407521, 0.466367027489468, 0.991560421418399, 0.883195955539122, 0.286449406063184, 0.973160334629938, 0.686454448383301, 0.0398739664815366, 0.669622014043853, 0.503212024457753, 0.015183342853561, 0.510261648334563, 0.595324236201122, 0.565353298094124, 0.283181632403284, 0.773779076291248, 0.768452184274793, 0.324779874412343, 0.583865500986576, 0.453149125212803, 0.0954202630091459, 0.882131988415495, 0.531715120188892, 0.173478826647624, 0.00553932529874146, 0.777186732040718, 0.95650587650016, 0.349630171898752, 0.469500013394281, 0.773521278984845, 0.5183498384431, 0.0248219578061253, 0.724037705687806, 0.0312377491500229, 0.559026331407949, 0.558551891241223, 0.218632768373936, 0.0496932289097458, 0.272120704408735, 0.369450616650283, 0.0713537756819278, 0.654042702168226, 0.706408486468717, 0.843981787096709, 0.325008332496509, 0.276549856178463, 0.327602964360267, 0.179422325454652, 0.566181435948238, 0.858145464211702, 0.601772548630834, 0.347887515788898, 0.809090912807733, 0.585183555260301, 0.859101112466305, 0.459006545133889, 0.487285570940003, 0.781371382065117, 0.571749424794689, 0.322488804114982, 0.0101655330508947, 0.569518544245511, 0.981841041706502, 0.747916206018999, 0.975305684609339, 0.176119895419106, 0.397235482698306, 0.36328464304097, 0.645597419235855, 0.798656917875633, 0.246906779939309, 0.907540212618187, 0.870592946186662, 0.0922842440195382, 0.700706091010943, 0.224282755982131, 0.124246483901516, 0.855422941036522, 0.15762529335916, 0.0353541523218155, 0.246808006428182, 0.097644732799381, 0.103977187070996, 0.272709888406098, 0.780843938700855, 0.637636936968192, 0.102873310446739, 0.814426307100803, 0.818499282002449, 0.533843900775537, 0.727865930413827, 0.432542552240193, 0.918857081793249, 0.443812903016806, 0.396129819797352, 0.385084046283737, 0.75622325623408, 0.279138972749934, 0.581082041142508, 0.541011241730303, 0.992757045198232, 0.237556868465617, 0.0896138437092304, 0.290659891208634, 0.557859183987603, 0.555533354869112, 0.237735997186974, 0.528231043368578, 0.421471484936774, 0.300639719469473, 0.900870520388708, 0.146541552385315, 0.741111336508766, 0.472619667649269, 0.0187025126069784, 0.963153589051217, 0.0432798964902759, 0.985922464868054, 0.780148433288559, 0.615160375833511, 0.978800310753286, 0.24654924031347, 0.172602775972337, 0.229555260855705, 0.398810408543795, 0.363575607305393, 0.749094863189384, 0.876136147417128, 0.752553601516411, 0.737650994211435, 0.714659651042894, 0.115543003426865, 0.934868972515687, 0.70222750515677, 0.909277478698641, 0.331805474823341, 0.181661519687623, 0.431351432111114, 0.924020344391465, 0.347989107016474, 0.46955297072418, 0.778679148294032, 0.7936415693257, 0.615784144727513, 0.00622527301311493, 0.590799591736868, 0.450871390523389, 0.298359509091824, 0.466935755452141, 0.412974033504725, 0.534691520966589, 0.714078595163301, 0.383068545488641, 0.118782989215106, 0.631121209356934, 0.597904303343967, 0.570675171911716, 0.115923219826072, 0.0469014558475465, 0.513423261465505, 0.390789948869497, 0.507823542691767, 0.191148014506325, 0.169754519127309, 0.41813335660845, 0.344705219147727, 0.902372638694942, 0.125337106408551, 0.679444398498163, 0.356002041837201, 0.849719845689833, 0.830207100370899, 0.674006941262633, 0.351869232952595, 0.559378106147051, 0.370021741604432, 0.0241289546247572, 0.235599905718118, 0.129839594243094, 0.736559700919315, 0.384421338094398, 0.991702135419473, 0.228056449210271, 0.647314116591588, 0.35237587033771, 0.123974025016651, 0.443579137092456, 0.449192600790411, 0.560574092436582, 0.289995907573029, 0.689897746080533, 0.071731336414814, 0.883814787724987, 0.942212292924523, 0.260529879946262, 0.716532932361588, 0.316868973197415, 0.722948840120807, 0.772855468094349, 0.875697809271514, 0.76832310971804, 0.450222017941996, 0.539084741147235, 0.782060565194115, 0.762430806644261, 0.284611386712641, 0.0379048513714224, 0.884668768616393, 0.964230293175206, 0.447210259037092, 0.544494225876406, 0.315904656657949, 0.0496735929045826, 0.453966638771817, 0.0809071147814393, 0.921034719096497, 0.188724118983373, 0.0537364124320447, 0.205119320191443, 0.736644972348586, 0.358727745711803, 0.0374156795442104, 0.669449704699218, 0.349910696269944, 0.826142649864778, 0.333134037908167, 0.786557076731697, 0.98594150878489, 0.897947886260226, 0.152906849980354, 0.823585299542174, 0.175882948562503, 0.0610524078365415, 0.447925634682178, 0.208358955103904, 0.720473872032017, 0.334872581064701, 0.979855271056294, 0.92368355137296, 0.367354336194694, 0.629088162910193, 0.838142207125202, 0.864915602374822, 0.103348140604794, 0.760919131804258, 0.716578650521114, 0.569822284393013, 0.454100054455921, 0.350065787788481, 0.581630904925987, 0.654333251528442, 0.33325978461653, 0.877372861141339, 0.954466701019555, 0.732069323537871, 0.712191378464922, 0.51160002942197, 0.254377734148875, 0.139921847730875, 0.914348456310108, 0.548160315025598, 0.101459727156907, 0.859762959647924, 0.145510538015515, 0.413285078248009, 0.123961373930797, 0.506698541343212, 0.372902704169974, 0.866836439818144, 0.00840391172096133, 0.576217962661758, 0.149433226324618, 0.601452287053689, 0.958165465388447, 0.480476718163118, 0.713979457272217, 0.038136029150337, 0.257292360300198, 0.950676222098991, 0.281128635630012, 0.425785727100447, 0.0818620112258941, 0.742464848794043, 0.0197805431671441, 0.712775551714003, 0.488908573286608, 0.441582834348083, 0.692185449181125, 0.899554135510698, 0.00568815087899566, 0.553238376276568, 0.460822017630562, 0.366132561583072, 0.508787144208327, 0.262514994945377, 0.984113744925708, 0.337186914635822, 0.362328284187242, 0.640446075471118, 0.690439167665318, 0.888694568304345, 0.0715998224914074, 0.285083642229438, 0.675709908129647, 0.207205336773768, 0.779416073812172, 0.874604681739584, 0.274111510254443, 0.816011066082865, 0.368276265915483, 0.0910361893475056, 0.172772919991985, 0.309004415059462, 0.522004497470334, 0.310291458852589, 0.935460967477411, 0.143955788575113, 0.171144550200552, 0.530260861385614, 0.361995994811878, 0.0223627530504018, 0.251268296735361, 0.224409508751705, 0.477795456303284, 0.648408759152517, 0.604998983209953, 0.643719711573794, 0.881748712621629, 0.154766360996291, 0.813729036366567, 0.175090042874217, 0.816938740899786, 0.0689233099110425, 0.359778376063332, 0.306504796724766, 0.380337670445442, 0.857221704907715, 0.0433378261514008, 0.612204012228176, 0.0219281136523932, 0.280201885383576, 0.820075199939311, 0.384033706970513, 0.311102657346055, 0.436969012022018, 0.702133365673944, 0.21537156496197, 0.134176019346341, 0.669569396413863, 0.724535046843812, 0.961253642803058, 0.131185678299516, 0.254775772802532, 0.528106466401368, 0.353068707510829, 0.821093623526394, 0.233737293398008, 0.776456519728526, 0.911426509730518, 0.434573647333309, 0.421660708496347, 0.540435385657474, 0.330485601909459, 0.333461952162907, 0.869136864319444, 0.766159744467586, 0.932038061786443, 0.550978524144739, 0.566929257940501, 0.862548047676682, 0.286349087720737, 0.568639802746475, 0.291920855874196, 0.939328803448007, 0.670871011912823, 0.35408042371273, 0.638017787598073, 0.322301434120163, 0.0179480123333633, 0.562558931298554, 0.851859271293506, 0.253767121583223, 0.0734880093950778, 0.399642747594044, 0.293491043383256, 0.123202317394316, 0.058303042082116, 0.803245912538841, 0.508685543434694, 0.586687728762627, 0.393899171845987, 0.062371889827773, 0.514366849791259, 0.432971445610747, 0.23642969597131, 0.237376803532243, 0.819560660747811, 0.105362235335633, 0.779521638061851, 0.319189456989989, 0.701739897951484, 0.106854846235365, 0.162746323505417, 0.88817330589518, 0.968502965057269, 0.740807080408558, 0.730465295957401, 0.938631818396971, 0.0468996313866228, 0.404501337558031, 0.766174633288756, 0.171461830846965, 0.325041643576697, 0.524244742002338, 0.105307154590264, 0.667091010604054, 0.0219771224074066, 0.498172964435071, 0.66492285951972, 0.952624740777537, 0.325426982948557, 0.569926236290485, 0.675628469092771, 0.199369940906763, 0.2875742088072, 0.347600874723867, 0.866687843343243, 0.941786771407351, 0.268384084571153, 0.121187934884802, 0.815922977402806, 0.0582089591771364, 0.939407591708004, 0.687808850547299, 0.460545796900988, 0.927214352414012, 0.159762202994898, 0.906076698796824, 0.425094907404855, 0.181808969005942, 0.0762422906700522, 0.731856314232573, 0.62383958697319, 0.988547759829089, 0.62334821629338, 0.313671520445496, 0.201175077585503, 0.513074612244964, 0.887830370105803, 0.0339405476115644, 0.327579334843904, 0.0272558457218111, 0.00610948237590492, 0.331780990818515, 0.162022365722805, 0.692474717041478, 0.472517163027078, 0.396922193467617, 0.16871378198266, 0.996896346332505, 0.889909054152668, 0.787555685266852, 0.502608251292259, 0.54986095125787, 0.0420358721166849, 0.746028373716399, 0.834122750908136, 0.242601352045313, 0.238337026443332, 0.219042910961434, 0.665373098105192, 0.84991412400268, 0.858990486711264, 0.383569119498134, 0.541636841371655, 0.534390223445371, 0.8340761102736, 0.322351802140474, 0.472435349831358, 0.825171621283516, 0.28880974301137, 0.31317821610719, 0.425488634966314, 0.0509746442548931, 0.560418299399316, 0.615369844250381, 0.813049888703972, 0.610890027368441, 0.704311636276543, 0.054966060211882, 0.679563012905419, 0.294183069141582, 0.165822205133736, 0.0527141510974616, 0.857890394283459, 0.471598664997146, 0.332361556356773, 0.816324507119134, 0.516102523542941, 0.578787426231429, 0.139616628410295, 0.688051319448277, 0.0792960757389665, 0.27170032938011, 0.559470106381923, 0.14075337187387, 0.237253423314542, 0.996158340480179, 0.737346267327666, 0.270884183701128, 0.621428779792041, 0.207560179755092, 0.290613756282255, 0.0982533732894808, 0.19001590134576, 0.487919705221429, 0.3891339560505, 0.405173858627677, 0.661630472401157, 0.109770814189687, 0.793682185234502, 0.00792052200995386, 0.400612744502723, 0.828204827383161, 0.378191909519956, 0.626916074659675, 0.496869592927396, 0.248783138114959, 0.931049096863717, 0.434795839013532, 0.24206828745082, 0.937603077851236, 0.710168126039207, 0.639684492256492, 0.662060776725411, 0.102620420977473, 0.4193098268006, 0.164752066601068, 0.329191894968972, 0.026295363670215, 0.368435434764251, 0.129059018800035, 0.909263829234987, 0.781660998007283, 0.446930978214368, 0.221178410109133, 0.354408835060894, 0.289195395773277, 0.531240687705576, 0.40554724750109, 0.647184815024957, 0.643816772382706, 0.100893170107156, 0.414878642419353, 0.658400003798306, 0.0681135824415833, 0.684170212596655, 0.480494690127671, 0.690349728567526, 0.470414221519604, 0.0805130687076598, 0.251933983992785, 0.182285729330033, 0.956523204222322, 0.311288903933018, 0.438309320947155, 0.2671473918017, 0.602850520052016, 0.00222030025906861, 0.851436564000323, 0.653805781621486, 0.327411290956661, 0.945507103810087, 0.0428934532683343, 0.892808785894886, 0.298424021806568, 0.627608072245494, 0.0119263101369143, 0.386439106194302, 0.0805595526471734, 0.00757312634959817, 0.155197300948203, 0.762971888296306, 0.38035902264528, 0.960596561664715, 0.772529227659106, 0.232518748613074, 0.306001355638728, 0.791862854734063, 0.624311193358153, 0.90246876468882, 0.112019885564223, 0.51626761094667, 0.140547572402284, 0.591676430311054, 0.700423401314765, 0.869920731522143, 0.653606876032427, 0.0400482369586825, 0.401266740635037, 0.0468274638988078, 0.953998438781127, 0.320154560729861, 0.475776583654806, 0.95948574366048, 0.598697250476107, 0.252777661662549, 0.911177820526063, 0.922542226267979, 0.0712300268933177, 0.986675314372405, 0.960889756679535, 0.608074870193377, 0.733704640064389, 0.525904290145263, 0.710961656877771, 0.669195500435308, 0.0614037208724767, 0.0302184249740094, 0.500243731308728, 0.0584362738300115, 0.822914346354082, 0.615969242528081, 0.670261591440067, 0.0648595322854817, 0.758059150772169, 0.645665595773607, 0.542095222510397, 0.468963160878047, 0.253484061453491, 0.264407187933102, 0.0916396786924452, 0.497509995708242, 0.447942925849929, 0.813280432252213, 0.979331208392978, 0.0546543973032385, 0.108002501307055, 0.746989071369171, 0.487027541035786, 0.611288634128869, 0.59845721302554, 0.341110674198717, 0.446535823866725, 0.62437161593698, 0.0701461411081254, 0.864706572145224, 0.508357064099982, 0.356072329683229, 0.846072605345398, 0.123709111707285, 0.866541317431256, 0.679472028277814, 0.0865811607800424, 0.199196594068781, 0.418080654228106, 0.188627138268203, 0.226113744778559, 0.260029500117525, 0.0399316335096955, 0.650983575964347, 0.19800451816991, 0.0255183689296246, 0.366445788415149, 0.551253109006211, 0.481654884526506, 0.578760679345578, 0.202811040682718, 0.456183291040361, 0.878565251128748, 0.203479228774086, 0.608080768026412, 0.532672078115866, 0.940779979107901, 0.414028265979141, 0.184072277741507, 0.0703828965779394, 0.117855264805257, 0.199670911999419, 0.747787687927485, 0.0776140480302274, 0.34225592110306, 0.964884238084778, 0.964255526429042, 0.577958170790225, 0.663519106339663, 0.0839966835919768, 0.92471485468559, 0.054384418996051, 0.260524642653763, 0.402208246523514, 0.585939414100721, 0.0293237986043096, 0.588579104747623, 0.431679262313992, 0.465528066037223, 0.98019934608601, 0.235518837813288, 0.879279229324311, 0.0205704206600785, 0.937828381545842, 0.0426210961304605, 0.813133704708889, 0.257024510065094, 0.872497952077538, 0.867703770054504, 0.80465923785232, 0.522447892930359, 0.049781059846282, 0.0114028954412788, 0.782095722621307, 0.743752285838127, 0.769396420801058, 0.108475690241903, 0.894451821688563, 0.830220413627103, 0.185815728269517, 0.265939095756039, 0.260238980408758, 0.72650664136745, 0.914975832682103, 0.389163572574034, 0.0530192023143172, 0.593306473921984, 0.896800601389259, 0.661566518712789, 0.975579047808424, 0.901407039025798, 0.206443930277601, 0.419750312110409, 0.78689140919596, 0.798239971743897, 0.692319952649996, 0.0247984013985842, 0.857957189204171, 0.501637978712097, 0.388112785527483, 0.0958032845519483, 0.0045508926268667, 0.0541927276644856, 0.733874855795875, 0.31278393487446, 0.357920822687447, 0.444655833765864, 0.581121385795996, 0.592959355562925, 0.301177168264985, 0.34394295909442, 0.368175566894934, 0.602365133119747, 0.312490853713825, 0.186573053477332, 0.938391978852451, 0.50495598372072, 0.665044444147497, 0.326559635344893, 0.329093999229372, 0.390392345376313, 0.891208411427215, 0.0774592526722699, 0.540256817359477, 0.912029358558357, 0.618063597241417, 0.604708577971905, 0.119590278482065, 0.904813922708854, 0.654458702309057, 0.908672583987936, 0.647809221874923, 0.381672145798802, 0.561931320698932, 0.943480239715427, 0.404441983671859, 0.628618631279096, 0.048962787957862, 0.00220109522342682, 0.467996649444103, 0.85375214763917, 0.0406711734831333, 0.460589741822332, 0.620278425514698, 0.0170565703883767, 0.971417004009709, 0.779673482989892, 0.254188556689769, 0.702849731082097, 0.345342008629814, 0.85889451415278, 0.831949535058811, 0.357516808900982, 0.821836359566078, 0.728659681975842, 0.22847291873768, 0.297777394531295, 0.247627233387902, 0.943180060014129, 0.594748799456283, 0.36701562977396, 0.842646370176226, 0.75786098651588, 0.363946038531139, 0.577017608098686, 0.576754971873015, 0.939791908953339, 0.119550654198974, 0.340325520606712, 0.398191246669739, 0.912185499910265, 0.348895763512701, 0.703890188131481, 0.605509347980842, 0.991284989286214, 0.47757355356589, 0.342707401374355, 0.371817533392459, 0.783800924196839, 0.833675854373723, 0.161071388982236, 0.331009736983106, 0.825950135942549, 0.15964378323406, 0.440975310979411, 0.823098176624626, 0.85917716473341, 0.775942172156647, 0.919696998316795, 0.333959904732183, 0.26250347844325, 0.517289029899985, 0.0919352862983942, 0.416853982955217, 0.709609117358923, 0.0901238776277751, 0.986764123197645, 0.061896649422124, 0.761099418159574, 0.804242689162493, 0.979894444579259, 0.929113615071401, 0.464212313992903, 0.384093690663576, 0.663515644846484, 0.907421685755253, 0.83422270603478, 0.602890911512077, 0.84451136784628, 0.271246948512271, 0.68092513945885, 0.13341801892966, 0.280375093687326, 0.409162634052336, 0.741109237773344, 0.148676952114329, 0.415354089112952, 0.775118483928964, 0.155097920913249, 0.598933037836105, 0.489613801706582, 0.960961676901206, 0.737901026615873, 0.396040980704129, 0.82669277000241, 0.365158501081169, 0.517252881778404, 0.050435149576515, 0.131452899891883, 0.293587620602921, 0.643413025420159, 0.285724128363654, 0.681313992710784, 0.381969437003136, 0.293921922566369, 0.388309161411598, 0.157212823163718, 0.51635954156518, 0.127340577309951, 0.763042234349996, 0.988424743758515, 0.695875012781471, 0.674544195877388, 0.638067554915324, 0.332506016595289, 0.0200457845348865, 0.721241177525371, 0.00871789711527526, 0.780132214538753, 0.609023758443072, 0.335371210472658, 0.28436201158911, 0.231810273602605, 0.754667499568313, 0.802990465657786), actual = c(1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.491785136051476, 0.850167193915695, 0.811472615227103, 0.800848407670856, 0.657440338749439, 0.0725702568888664, 0.319772005546838, 0.944550709333271, 0.205791883869097, 0.049539600731805, 0.0249058520421386, 0.339855366153643, 0.703011644305661, 0.376338495407254, 0.545719145331532, 0.217725543538108, 0.385697247460485, 0.361176915233955, 0.0724715967662632, 0.637563202762976, 0.110080844024196, 0.665290563832968, 0.832701381994411, 0.0702630029991269, 0.731409357860684, 0.383683307096362, 0.0181832986418158, 0.892846883973107, 0.184077375335619, 0.448036837624386, 0.634507367853075, 0.724210700020194, 0.378318496281281, 0.0743342007044703, 0.135455636540428, 0.041373826097697, 0.153519130311906, 0.263803546782583, 0.802861720556393, 0.424060781486332, 0.971401676535606, 0.16156910918653, 0.538562550442293, 0.409966791747138, 0.862241113791242, 0.956977220484987, 0.835334942210466, 0.948050356470048, 0.391686738003045, 0.458140904316679, 0.845709729008377, 0.787666505901143, 0.298341055167839, 0.089429097250104, 0.574180574854836, 0.74632142437622, 0.0329707092605531, 0.468573142541572, 0.685179763939232, 0.841755213215947, 0.280489669181406, 0.337956012925133, 0.924895465141162, 0.805915864882991, 0.842390851583332, 0.873456079745665, 0.370595755055547, 0.50780518213287, 0.587803855305538, 0.260681271320209, 0.323472516611218, 0.809985539177433, 0.880101955030113, 0.491299451561645, 0.951467214850709, 0.21910348511301, 0.712315892567858, 0.931955608073622, 0.390408256556839, 0.822947431355715, 0.380158650921658, 0.513359872391447, 0.614171445835382, 0.313406591536477, 0.854046052787453, 0.342436421196908, 0.0751321699935943, 0.465175548102707, 0.087328887777403, 0.430741578806192, 0.0440605599433184, 0.396923368331045, 0.911042171996087, 0.403006772743538, 0.965007181745023, 0.66731513175182, 0.617280861362815, 0.429527839878574, 0.645001946017146, 0.836743367370218, 0.704020133242011, 0.965107619995251, 0.0725152282975614, 0.775359983323142, 0.768310821149498, 0.400109707377851, 0.0872502592392266, 0.889712158823386, 0.314302455401048, 0.204595641698688, 0.206114841159433, 0.981606512563303, 0.609589726198465, 0.0994057692587376, 0.985156817128882, 0.584432408679277, 0.447072985349223, 0.897094969172031, 0.44899511593394, 0.262835095869377, 0.336631597485393, 0.764933547005057, 0.609986326890066, 0.0837757217232138, 0.0180210878606886, 0.0479728823993355, 0.908975149970502, 0.682927400572225, 0.0927265619393438, 0.293128608260304, 0.672439897432923, 0.493225800339133, 0.182207875885069, 0.403513286495581, 0.186119511490688, 0.915848941775039, 0.214633714640513, 0.0149456509388983, 0.841177860973403, 0.180753772612661, 0.543777173850685, 0.849287044955418, 0.449034270597622, 0.719078466063365, 0.574763340409845, 0.971949385479093, 0.459624911425635, 0.758990751113743, 0.74183266190812, 0.129374178126454, 0.637238607974723, 0.336127489572391, 0.260572370374575, 0.223027570638806, 0.647586462087929, 0.320430578896776, 0.90676175057888, 0.486827410990372, 0.410293017979711, 0.58132625091821, 0.209009232232347, 0.0190083829220384, 0.859959752531722, 0.797163898823783, 0.375406129984185, 0.715132065117359, 0.435333855682984, 0.231575938640162, 0.428460583556443, 0.856176676694304, 0.0241826951969415, 0.853159131715074, 0.672248564660549, 0.511926095932722, 0.668521228944883, 0.21327258925885, 0.103120270883664, 0.741775603266433, 0.695191423641518, 0.119947399478406, 0.805606764508411, 0.0861223791725934, 0.712817221879959, 0.92876533488743, 0.0519530253950506, 0.419065796770155, 0.379235542844981, 0.60014727874659, 0.525368739617988, 0.0825798879377544, 0.0685146327596158, 0.191064472077414, 0.495323363691568, 0.894773501902819, 0.797868135385215, 0.689309284556657, 0.123784227762371, 0.766406286275014, 0.37403028504923, 0.33206335012801, 0.121824314352125, 0.792365914909169, 0.712552543263882, 0.472496943315491, 0.657697976799682, 0.894872085191309, 0.723048365209252, 0.46316338586621, 0.500020327046514, 0.857638146961108, 0.455235075904056, 0.989762155804783, 0.179281025193632, 0.613368287915364, 0.979518132982776, 0.0735361324623227, 0.5261386774946, 0.1172866798006, 0.694149974966422, 0.232003407785669, 0.334885384654626, 0.942399158142507, 0.195711625739932, 0.165692468173802, 0.46535033849068, 0.827698109904304, 0.770851955749094, 0.411785229574889, 0.195409850915894, 0.182985532097518, 0.284923551604152, 0.81642533512786, 0.784486924065277, 0.951383684761822, 0.42996061523445, 0.67604753584601, 0.0749920017551631, 0.922786653507501, 0.729476145468652, 0.062440438196063, 0.994083150988445, 0.678140510572121, 0.429636431392282, 0.28919740836136, 0.725762563291937, 0.61397404759191, 0.488825417589396, 0.263092420529574, 0.246872571995482, 0.927161256317049, 0.766867840662599, 0.951372314011678, 0.544841029448435, 0.734102072892711, 0.794331951066852, 0.519073394360021, 0.191421719966456, 0.362341623287648, 0.581392902880907, 0.605599394068122, 0.279359956504777, 0.256045736838132, 0.320797143969685, 0.180334841366857, 0.954337167786434, 0.469195758923888, 0.323956721229479, 0.694345762953162, 0.45013769925572, 0.376738880993798, 0.122987450566143, 0.467148360796273, 0.248251876560971, 0.738817447796464, 0.0622723100241274, 0.67055109818466, 0.503466891357675, 0.227994114160538, 0.593730011489242, 0.31399890081957, 0.157832014840096, 0.963509347289801, 0.299657639348879, 0.959247052669525, 0.652641211403534, 0.129876944236457, 0.483570852549747, 0.551746709737927, 0.132646074518561, 0.617262481246144, 0.133351973956451, 0.79948748787865, 0.317397930426523, 0.0902686724439263, 0.654801873257384, 0.234744232147932, 0.793128564022481, 0.0319867732468992, 0.988505269633606, 0.237745781429112, 0.91571413166821, 0.387874922249466, 0.645814823685214, 0.207432542461902, 0.503397299908102, 0.363995151827112, 0.353393164230511, 0.322635892545804, 0.118000527145341, 0.649994072970003, 0.183778495294973, 0.82430919748731, 0.901067229220644, 0.559765031561255, 0.578552854945883, 0.288499878253788, 0.297831220552325, 0.326014279155061, 0.439926970750093, 0.771396908676252, 0.82183611090295, 0.842588514089584, 0.358811313286424, 0.725486809387803, 0.196569004561752, 0.670415300643072, 0.578002169029787, 0.76652827183716, 0.598086901940405, 0.363506661728024, 0.276770212454721, 0.369323245016858, 0.425688483286649, 0.232122464105487, 0.23182886838913, 0.363820557482541, 0.173227226128802, 0.952228688634932, 0.156289563281462, 0.3448927460704, 0.157169216312468, 0.504067908041179, 0.49684536177665, 0.845709329238161, 0.24592663324438, 0.854470956604928, 0.528893710812554, 0.963659671833739, 0.0457653927151114, 0.853951656958088, 0.612530199112371, 0.641183773055673, 0.142841936321929, 0.999216330004856, 0.0726787014864385, 0.33427930646576, 0.502305397763848, 0.917434160131961, 0.806393791921437, 0.731501789297909, 0.990348312538117, 0.669218771858141, 0.115898788208142, 0.006336237071082, 0.667765831807628, 0.164678113767877, 0.074413834605366, 0.269702889723703, 0.204924599733204, 0.480807108106092, 0.123990494292229, 0.682751424610615, 0.311248453101143, 0.54701939993538, 0.868257729103789, 0.692175471223891, 0.746824832633138, 0.965302078984678, 0.615926091326401, 0.744787514908239, 0.534105845727026, 0.589135941350833, 0.506888593314216, 0.99126471253112, 0.246021174592897, 0.607110392535105, 0.192469119327143, 0.480918164597824, 0.951506689190865, 0.604041539831087, 0.123367729363963, 0.618254687637091, 0.0675216373056173, 0.501755048986524, 0.592641200870275, 0.336533010704443, 0.802454200573266, 0.870439135236666, 0.129233670420945, 0.075170048745349, 0.618910846067593, 0.13644285663031, 0.363137453794479, 0.0344637567177415, 0.109654451021925, 0.119784083683044, 0.610329047543928, 0.0491988973226398, 0.729519220534712, 0.594961499562487, 0.638871649513021, 0.624178410740569, 0.362540584756061, 0.9347288091667, 0.0698130035307258, 0.497884716838598, 0.156246605562046, 0.416001002769917, 0.119024037150666, 0.186647959286347, 0.933278202079237, 0.517590580275282, 0.477942594094202, 0.628770819166675, 0.833178917644545, 0.0944875713903457, 0.540391584625468, 0.501522904494777, 0.939000269863755, 0.504057519836351, 0.956174621824175, 0.117430760757998, 0.597545026568696, 0.585414032218978, 0.0523917623795569, 0.792056048056111, 0.741319760214537, 0.655167730757967, 0.617389609338716, 0.444810496643186, 0.620100975269452, 0.0916525751817971, 0.158620293717831, 0.457968012895435, 0.0315092012751848, 0.010842083254829, 0.465394066181034, 0.574243466136977, 0.80753688653931, 0.192545785335824, 0.87223436567001, 0.936059460509568, 0.682714097900316, 0.00960144470445812, 0.56586331105791, 0.522939050802961, 0.773007821990177, 0.781885346863419, 0.185656081419438, 0.0818600058555603, 0.795588795561343, 0.261383153265342, 0.661626093555242, 0.725566613255069, 0.480961043853313, 0.72960017202422, 0.0859035192988813, 0.0778832107316703, 0.720020878361538, 0.443240812048316, 0.084821430966258, 0.881914591416717, 0.91953266877681, 0.301031248411164, 0.389401882421225, 0.684558923356235, 0.846382503863424, 0.670376121299341, 0.0492910789325833, 0.358047847170383, 0.839759419672191, 0.161739689763635, 0.221110324608162, 0.919277776498348, 0.891393332974985, 0.963154154829681, 0.514819381060079, 0.929418836487457, 0.585999279748648, 0.138841692591086, 0.532356120413169, 0.371151886880398, 0.909427522215992, 0.897735962178558, 0.334771151421592, 0.855408747214824, 0.13836512202397, 0.014204022474587, 0.813830974046141, 0.0274802434723824, 0.393699329346418, 0.86776059307158, 0.509679649258032, 0.823489969829097, 0.617104531032965, 0.876200562808663, 0.532360739773139, 0.122142625739798, 0.292963956017047, 0.377032715361565, 0.651171454694122, 0.0872383492533118, 0.595080401748419, 0.525650252122432, 0.407778856344521, 0.351069102296606, 0.407715568551794, 0.334902463946491, 0.788320976542309, 0.115381792187691, 0.198188483482227, 0.197061450919136, 0.618345683673397, 0.651216187980026, 0.865496719954535, 0.292968517635018, 0.836723410291597, 0.940157862380147, 0.925606416538358, 0.79037652304396, 0.493899375665933, 0.999435024568811, 0.0930854508187622, 0.612351869000122, 0.524674959480762, 0.679481491912156, 0.616266228957102, 0.253665965050459, 0.487289972836152, 0.79080531722866, 0.679811068577692, 0.380738819250837, 0.130845929961652, 0.293172761565074, 0.625077343545854, 0.209991082083434, 0.393959082430229, 0.00897667650133371, 0.75001382175833, 0.949112562928349, 0.78495844360441, 0.712422850774601, 0.749592064647004, 0.453905766364187, 0.165376254357398, 0.575671455590054, 0.15213833283633, 0.542780499206856, 0.101451015332714, 0.896063133142889, 0.970978695899248, 0.799234360456467, 0.916298644850031, 0.20264262962155, 0.211557060014457, 0.126093421131372, 0.325051756342873, 0.314986299956217, 0.472160377074033, 0.123472034698352, 0.891369147924706, 0.255803027190268, 0.482291962020099, 0.388346292776987, 0.788791500497609, 0.54921634029597, 0.225863474188372, 0.224352581659332, 0.28346210392192, 0.514457334997132, 0.81901517463848, 0.353355712722987, 0.989147602347657, 0.613059048773721, 0.0467401342466474, 0.561456674942747, 0.363492907956243, 0.809592451201752, 0.849931033095345, 0.277056929888204, 0.758305336115882, 0.752675563097, 0.399601578246802, 0.483286510920152, 0.0432059417944402, 0.764549024635926, 0.737196902744472, 0.0876654537860304, 0.717139539308846, 0.587767815217376, 0.854556877631694, 0.965869665844366, 0.128739100182429, 0.670876655029133, 0.50376713857986, 0.817968135233968, 0.751574340974912, 0.091523775132373, 0.395391189260408, 0.687603452475742, 0.493798793526366, 0.874695033533499, 0.502915048738942, 0.856733188033104, 0.0336410475429147, 0.228841856587678, 0.387273256666958, 0.205341409426183, 0.718259279616177, 0.0355529859662056, 0.999012947781011, 0.705724040977657, 0.925378929125145, 0.0712241935543716, 0.608696985058486, 0.67229858902283, 0.873048645909876, 0.00387211516499519, 0.973101949552074, 0.918667021207511, 0.453936416190118, 0.913322030100971, 0.862303905421868, 0.906461234437302, 0.929437801707536, 0.654290035367012, 0.713342186994851, 0.273441642755643, 0.116436989512295, 0.74598405812867, 0.930278223240748, 0.643000797834247, 0.882420015288517, 0.110178945120424, 0.508657502243295, 0.933289590058848, 0.148220349801704, 0.941508457995951, 0.407302354462445, 0.309394729556516, 0.634036118164659, 0.96460337145254, 0.0909263791982085, 0.967999736079946, 0.952900543110445, 0.407281881431118, 0.0592963960953057, 0.533122697146609, 0.142733499640599, 0.916302371537313, 0.494437920628116, 0.954224458429962, 0.522602444048971, 0.360662975814193, 0.571595248999074, 0.777965022483841, 0.0187742519192398, 0.0960352497640997, 0.00757326139137149, 0.181310744956136, 0.245980392908677, 0.339170254068449, 0.162983563262969, 0.575427810195833, 0.565423636930063, 0.335249511757866, 0.608324465807527, 0.449738196330145, 0.32855529198423, 0.225436682347208, 0.184716258663684, 0.617664119461551, 0.0138404439203441, 0.516786741558462, 0.386727229459211, 0.367942120181397, 0.29153328249231, 0.871887261746451, 0.406043425668031, 0.0857060162816197, 0.308351872954518, 0.454023796366528, 0.189902160316706, 0.951116977958009, 0.676317127654329, 0.143975824583322, 0.227295486489311, 0.535732260439545, 0.72269815672189, 0.0286643619183451, 0.855978108011186, 0.0222877627238631, 0.308945244876668, 0.307310229865834, 0.525929459370673, 0.342077788198367, 0.177528405794874, 0.937388687394559, 0.152708318317309, 0.187095554312691, 0.53589901025407, 0.0442200398538262, 0.0370223945938051, 0.5128377340734, 0.277559061301872, 0.789325832156464, 0.522789267124608, 0.554472248535603, 0.0315010219346732, 0.425331179751083, 0.97116209147498, 0.421412788331509, 0.85037515219301, 0.121891513234004, 0.726857884321362, 0.331461034948006, 0.664079756243154, 0.759581628255546, 0.921873607672751, 0.889558483380824, 0.804722360568121, 0.314733537146822, 0.719262819271535, 0.100861607585102, 0.430343626765534, 0.224012969760224, 0.986684228293598, 0.426984057761729, 0.0477203656919301, 0.986718343570828, 0.667306852992624, 0.484349235892296, 0.094184601213783, 0.698522257152945, 0.665434277616441, 0.355742981657386, 0.831018288386986, 0.110672431066632, 0.0487461350858212, 0.183388387318701, 0.373097833245993, 0.987831133883446, 0.935355762951076, 0.274338368559256, 0.577393157873303, 0.879405860323459, 0.66251063067466, 0.952970506623387, 0.460541444364935, 0.349913472542539, 0.747471985174343, 0.0240140561945736, 0.900445557432249, 0.0570272856857628, 0.155960679752752, 0.00842078775167465, 0.809307937743142, 0.923658252460882, 0.836147355381399, 0.105901151662692, 0.840676269261166, 0.120790242217481, 0.451410802546889, 0.874280190793797, 0.32007625605911, 0.519444569246843, 0.274817924015224, 0.103120102081448, 0.636488552438095, 0.00955296750180423, 0.700851992703974, 0.358723055105656, 0.961144543951377, 0.97668939223513, 0.24634819291532, 0.397905833786353, 0.817040194058791, 0.230746279936284, 0.345961090177298, 0.531510598491877, 0.841559343738481, 0.921550174010918, 0.37659843149595, 0.493021452566609, 0.879959465935826, 0.705238462425768, 0.648296405561268, 0.296166487038136, 0.866688173264265, 0.577044193632901, 0.758682176470757, 0.087707530008629, 0.964947112603113, 0.278758123284206, 0.728442615829408, 0.130526155931875, 0.322583002271131, 0.599526461912319, 0.317523318110034, 0.673520403681323, 0.343168759718537, 0.261702059069648, 0.372939347056672, 0.218582923291251, 0.78000227897428, 0.877713940804824, 0.759835877921432, 0.396765187382698, 0.114549953490496, 0.819352818187326, 0.504898380488157, 0.953797810478136, 0.918044440681115, 0.191857293713838, 0.887301207054406, 0.912559705786407, 0.183450507931411, 0.680130209075287, 0.72379256808199, 0.400129406014457, 0.516524501610547, 0.484102627960965, 0.931022290838882, 0.0711938499007374, 0.808737331070006, 0.0656059302855283, 0.865788775729015, 0.18480059155263, 0.791171551449224, 0.72576640220359, 0.0257403426803648, 0.761264979373664, 0.725200537359342, 0.907891811570153, 0.102459233719856, 0.0456554053816944, 0.47844817256555, 0.924118978669867, 0.846798091661185, 0.760878782253712, 0.638951450353488, 0.633725752355531, 0.176664967089891, 0.126904540229589, 0.742695472668856, 0.109591308515519, 0.181829519337043, 0.819281334290281, 0.562180108390749, 0.622836261056364, 0.350482495035976, 0.423199673416093, 0.383988687302917, 0.133612485369667, 0.547853532014415, 0.0180271861609071, 0.952711483929306, 0.55635429895483, 0.453494688495994, 0.957847216399387, 0.985011858167127, 0.157387035666034, 0.345484799472615, 0.438777494244277, 0.9584484428633, 0.579177702311426, 0.466782919829711, 0.735142323654145, 0.375798285007477, 0.0516962674446404, 0.983872960088775, 0.725081629352644, 0.200171679724008, 0.91135528520681, 0.49848201777786, 0.686387015739456, 0.750285558635369, 0.0304624871350825, 0.936497681075707, 0.97896928479895, 0.735136606264859, 0.906808879924938, 0.0430022904183716, 0.251318492926657, 0.485807810444385, 0.416092035826296, 0.403470847755671, 0.919190377928317, 0.460170263424516, 0.851489369524643, 0.963275633053854, 0.148384675616398, 0.05135658220388, 0.573523986618966, 0.00668235775083303, 0.601079138694331, 0.448889510240406, 0.994547829963267, 0.950328658567742, 0.705745532410219, 0.00547300139442086, 0.0931600274052471, 0.298174562165514, 0.491739789023995, 0.0815568957477808, 0.492253781761974, 0.592760692117736, 0.37606736086309, 0.325219975784421, 0.984360171249136, 0.0134142769966274, 0.708085627993569, 0.608730148524046, 0.446692230645567, 0.697722120443359, 0.780729229561985, 0.812030717497692, 0.678563981316984, 0.755099142435938, 0.685431137215346, 0.878294561058283, 0.885247872909531, 0.0544094629585743, 0.718700344208628, 0.89241357264109, 0.96307723573409, 0.508472781861201, 0.826770128449425, 0.886773013742641, 0.926722643198445, 0.33827150403522, 0.0260749459266663, 0.392697438364848, 0.95547122368589, 0.732804894912988, 0.699994337745011, 0.185263109160587, 0.851393678225577, 0.304159057559446, 0.421044879127294, 0.93041001399979, 0.346909790299833, 0.728728499496356, 0.720655984943733, 0.265835383441299, 0.025605124887079, 0.0188289694488049, 0.782447780482471, 0.226613866630942, 0.705357032595202, 0.984490604372695, 0.998017357196659, 0.846153754508123, 0.355121420230716, 0.819327117409557, 0.621676710201427, 0.529763151425868, 0.065770206740126, 0.945728497812524, 0.98419075878337, 0.358929448295385, 0.605881290975958, 0.749392882687971, 0.405090658692643, 0.965929872356355, 0.588464401196688, 0.264060580404475, 0.214495335472748, 0.0789450891315937, 0.872041385620832, 0.512910179328173, 0.336162189487368, 0.505588362459093, 0.399369292659685, 0.77417565183714, 0.202299368567765, 0.31160411098972, 0.89477091957815, 0.856472490821034, 0.719295764574781, 0.7623480563052, 0.149022420169786, 0.469250133493915, 0.265259576030076, 0.763134151697159, 0.494134078035131, 0.747567977989092, 0.945018119644374, 0.444906870368868, 0.0368605018593371), actual = c(0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"