Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.8079244        0.8079244 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5033335        0.5033335        0.5153874 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.216066473163664, 0.0453742542304099, 0.595346774207428, 0.218224604381248, 0.718727621249855, 0.278190961806104, 0.277951016556472, 0.220332919387147, 0.257253086660057, 0.856026537017897, 0.173520712414756, 0.964143997058272, 0.101212730398402, 0.427852662978694, 0.740458424203098, 0.327526674373075, 0.428962567821145, 0.967358531197533, 0.10720938607119, 0.465059185167775, 0.0104699006769806, 0.123287699883804, 0.348280789330602, 0.0164590084459633, 0.467550446279347, 0.691151259234175, 0.618798317853361, 0.0426978901959956, 0.352332206210122, 0.998805185314268, 0.756717969430611, 0.154799792217091, 0.419348560040817, 0.301977320341393, 0.160873444983736, 0.567024768330157, 0.931571206543595, 0.76398688624613, 0.393042124109343, 0.795023871818557, 0.410412661498412, 0.768125297501683, 0.26778329606168, 0.178020716179162, 0.355104073882103, 0.476461844053119, 0.724826114019379, 0.333050781162456, 0.279083021916449, 0.714403121266514, 0.09027004474774, 0.207432110793889, 0.997561487369239, 0.178826488321647, 0.410921971779317, 0.477762955706567, 0.516746654640883, 0.606673724250868, 0.887127150315791, 0.870378614403307, 0.341445329133421, 0.121601999737322, 0.392008689697832, 0.144273680169135, 0.826940618921071, 0.0371822360903025, 0.600165150361136, 0.802773643750697, 0.441722725983709, 0.194020482944325, 0.411753226304427, 0.969527844339609, 0.340852916473523, 0.519887266447768, 0.595995395909995, 0.297433812171221, 0.738292857771739, 0.0391787921544164, 0.656687568407506, 0.0805260115303099, 0.343830219469965, 0.226837370311841, 0.391799228265882, 0.267478779423982, 0.967218989972025, 0.317505282349885, 0.658388777403161, 0.10450642881915, 0.174918759614229, 0.986246778164059, 0.30691518750973, 0.823646297445521, 0.815303778275847, 0.149272511480376, 0.0402376782149076, 0.611240268452093, 0.64962458377704, 0.19022026960738, 0.571016321890056, 0.780883084284142, 0.79744967375882, 0.654380928259343, 0.624714750330895, 0.711565111530945, 0.920479451073334, 0.953299498418346, 0.641520468285307, 0.332107686437666, 0.208960324991494, 0.287217350676656, 0.0854514117818326, 0.182894558412954, 0.624702495988458, 0.436885760165751, 0.34587137424387, 0.817556487862021, 0.429320618975908, 0.305864343652502, 0.129226527176797, 0.979239150881767, 0.00819258741103113, 0.727214851649478, 0.291859874501824, 0.242421311093494, 0.251948228571564, 0.539802724728361, 0.454979246482253, 0.0668789681512862, 0.810070374747738, 0.586353442166001, 0.407463850686327, 0.481399699114263, 0.865448793396354, 0.400519003160298, 0.897153886267915, 0.0676216578576714, 0.851050696801394, 0.143703000620008, 0.197177625028417, 0.899738202802837, 0.141390628181398, 0.436904881615192, 0.443643886595964, 0.538008923875168, 0.496802550973371, 0.311739278491586, 0.136520940577611, 0.0435395515523851, 0.540862906957045, 0.128316993126646, 0.190369398100302, 0.55593046057038, 0.853324304334819, 0.964125667465851, 0.365409940248355, 0.349916343810037, 0.65523890289478, 0.577530588721856, 0.318564048968256, 0.30761576606892, 0.656093059573323, 0.667241998948157, 0.76462492113933, 0.484658762812614, 0.274820694467053, 0.359743661247194, 0.413094373419881, 0.903093090513721, 0.200939282542095, 0.65264014294371, 0.415213031228632, 0.935288812499493, 0.503511686110869, 0.945557095808908, 0.0673544844612479, 0.698237565811723, 0.789179275510833, 0.487522260285914, 0.181776764336973, 0.195746828336269, 0.923291935119778, 0.244361578952521, 0.451807609992102, 0.231119233649224, 0.595881818328053, 0.0508579039014876, 0.121669399552047, 0.25090275448747, 8.16390383988619e-05, 0.960844554938376, 0.678624009713531, 0.102697662077844, 0.827837479067966, 0.00634672329761088, 0.466240704292431, 0.327807195484638, 0.636874061077833, 0.104851241223514, 0.0517330202274024, 0.0117062248755246, 0.481473441235721, 0.129063468193635, 0.285944869741797, 0.635379875777289, 0.0173706314526498, 0.308270267909393, 0.663327543996274, 0.362796309171245, 0.221494409954175, 0.284399630036205, 0.145245969528332, 0.499313441338018, 0.0109464190900326, 0.733162978198379, 0.450269576627761, 0.368897921638563, 0.698191307485104, 0.425441858824342, 0.144156929338351, 0.0981948596891016, 0.779115590965375, 0.853433972224593, 0.677799094468355, 0.0772650418803096, 0.474805416539311, 0.394014894263819, 0.811427727108821, 0.588978136656806, 0.681986800394952, 0.273459737421945, 0.534483399940655, 0.821142326807603, 0.638819808140397, 0.820780585054308, 0.907798095839098, 0.591785793658346, 0.129045932320878, 0.578844573348761, 0.265167131554335, 0.0439286488108337, 0.0883431641850621, 0.497438551858068, 0.919827087083831, 0.33320536906831, 0.792825491866097, 0.379483122611418, 0.391353255370632, 0.754303369671106, 0.937307759420946, 0.659321998478845, 0.664895804366097, 0.04900634707883, 0.834755326388404, 0.0716022276319563, 0.637551169842482, 0.57686283160001, 0.19176855799742, 0.707379541127011, 0.430460229516029, 0.737972215283662, 0.220298850443214, 0.736701682442799, 0.233401272445917, 0.275016838684678, 0.51849466888234, 0.630921674426645, 0.385007897391915, 0.881341050378978, 0.706739152316004, 0.858790062833577, 0.511414978886023, 0.349778968840837, 0.0743542292620987, 0.725285112392157, 0.180887273047119, 0.461204573977739, 0.323966423980892, 0.455398553051054, 0.720140227349475, 0.915563391055912, 0.868711886694655, 0.248048295732588, 0.303954482544214, 0.92443358222954, 0.715713276527822, 0.289867799729109, 0.129833952058107, 0.511872324626893, 0.567851067287847, 0.236016241135076, 0.274270181078464, 0.445070918649435, 0.512632738100365, 0.22432930697687, 0.99305884516798, 0.331080998759717, 0.857651529135182, 0.602309389272705, 0.764749634545296, 0.443421909818426, 0.181542177684605, 0.707352385623381, 0.788876848993823, 0.884920330019668, 0.0415516048669815, 0.388104570098221, 0.41463488782756, 0.143187892623246, 0.400575530482456, 0.110439701704308, 0.444458460202441, 0.845428735250607, 0.238999813096598, 0.543970112688839, 0.151946384692565, 0.853613068349659, 0.309342858847231, 0.243086063535884, 0.0928004707675427, 0.816182938171551, 0.500091192312539, 0.685889485292137, 0.141295982990414, 0.561728757573292, 0.998218407388777, 0.895580127835274, 0.255262205842882, 0.51795866410248, 0.610725827049464, 0.902739392593503, 0.770190031267703, 0.910738032544032, 0.691675996407866, 0.772791097639129, 0.13889643503353, 0.0662277401424944, 0.834312648512423, 0.728042201371863, 0.0404160157777369, 0.88910640636459, 0.451063745189458, 0.892177826957777, 0.0196368319448084, 0.513187414733693, 0.76489764242433, 0.757723713992164, 0.624767116270959, 0.905310582369566, 0.491299055982381, 0.72758851153776, 0.797080953605473, 0.350145084550604, 0.504054214339703, 0.994546243455261, 0.0840067914687097, 0.518014801898971, 0.484530064743012, 0.358688137959689, 0.0776517651975155, 0.0692831110209227, 0.686421781778336, 0.446572290733457, 0.308803986757994, 0.831711834063753, 0.841483492869884, 0.570933434413746, 0.770440870895982, 0.263737894361839, 0.256639757659286, 0.350494367303327, 0.913070304086432, 0.0427467073313892, 0.630507859867066, 0.58050723047927, 0.773985966807231, 0.0735377720557153, 0.804557134862989, 0.216641191160306, 0.835154385771602, 0.898018692154437, 0.365411380073056, 0.828387306537479, 0.727288896450773, 0.810538338730112, 0.952336526941508, 0.825080475304276, 0.801820892607793, 0.751708319876343, 0.825544029008597, 0.87720913765952, 0.450010744621977, 0.610148837324232, 0.391808677697554, 0.809734255541116, 0.281656584236771, 0.293147220043465, 0.922161476919428, 0.96080585476011, 0.325398446526378, 0.327314670197666, 0.997273514978588, 0.270611210493371, 0.781144504901022, 0.983520117821172, 0.194874923909083, 0.56038887379691, 0.993497723247856, 0.542225712677464, 0.810884084319696, 0.292489589657634, 0.523554644314572, 0.741401600185782, 0.567126769805327, 0.472503432072699, 0.540259905857965, 0.96303614298813, 0.910704123554751, 0.281258320901543, 0.470205276506022, 0.546557760098949, 0.326748809311539, 0.0570862467866391, 0.974191883811727, 0.342013493878767, 0.444815562805161, 0.834429334616289, 0.785552560118958, 0.204202358610928, 0.60968039999716, 0.552489048335701, 0.494621754158288, 0.878672216553241, 0.315543056931347, 0.552312044193968, 0.965393503895029, 0.0646146233193576, 0.152665895177051, 0.441238577943295, 0.163808981887996, 0.755091892788187, 0.392037276411429, 0.796477410942316, 0.623282924294472, 0.317795333219692, 0.823665031697601, 0.573218257399276, 0.346946600126103, 0.974973914213479, 0.373078056611121, 0.448655205080286, 0.193103139288723, 0.408379255328327, 0.109165469650179, 0.715702960267663, 0.305726382182911, 0.311829045182094, 0.0517477579414845, 0.647218657890335, 0.783193485811353, 0.748340567341074, 0.875589643837884, 0.873730102321133, 0.156021344242617, 0.745855320710689, 0.381392251700163, 0.494569265749305, 0.332362042972818, 0.928121593547985, 0.136799190659076, 0.548709615832195, 0.853829051600769, 0.839206816628575, 0.57793131028302, 0.865015037823468, 0.0591934637632221, 0.790892236866057, 0.119082755874842, 0.544874424813315, 0.783937902189791, 0.806978504173458, 0.0419512575026602, 0.590938319917768, 0.209299218608066, 0.26927027804777, 0.878809116082266, 0.777055605780333, 0.186442360281944, 0.743422011379153, 0.989797513699159, 0.258126968983561, 0.410296120680869, 0.784316191682592, 0.539056942099705, 0.211740460712463, 0.62344567081891, 0.608328987378627, 0.334982326021418, 0.455776966409758, 0.0959377705585212, 0.547510172473267, 0.491222153417766, 0.96171150659211, 0.359516386641189, 0.942741010105237, 0.581382269505411, 0.313959246268496, 0.762970385840163, 0.908541492419317, 0.619597561191767, 0.0354428996797651, 0.536797147011384, 0.952196154044941, 0.78357241791673, 0.564682635944337, 0.907930494518951, 0.403177187778056, 0.728175026830286, 0.521695468341932, 0.500372508307919, 0.294708158355206, 0.690470298519358, 0.942295945016667, 0.582760601304471, 0.365467307157815, 0.02096206555143, 0.00522572477348149, 0.191349099157378, 0.916773521574214, 0.736009446438402, 0.903061629505828, 0.20842259330675, 0.239017441635951, 0.632703686831519, 0.870060725603253, 0.409116094000638, 0.157767296535894, 0.666768199298531, 0.421317139873281, 0.294157766038552, 0.969101299997419, 0.695423629600555, 0.0771569877397269, 0.466115715680644, 0.868888847995549, 0.139219446107745, 0.158120575826615, 0.00420258287340403, 0.206261760555208, 0.438307360745966, 0.229155987501144, 0.516888939309865, 0.509300947422162, 0.349886398296803, 0.138591320952401, 0.350207189563662, 0.0373600309249014, 0.876633062725887, 0.778523330111057, 0.693271507741883, 0.805974323535338, 0.0253198088612407, 0.586052644299343, 0.0256222663447261, 0.403890523361042, 0.0907417384441942, 0.792871977901086, 0.804929778911173, 0.956302504753694, 0.782502134563401, 0.374810559907928, 0.836658506188542, 0.24054117873311, 0.778036083560437, 0.869096446782351, 0.935666016070172, 0.93303930433467, 0.706189504824579, 0.726935734506696, 0.523455886403099, 0.293397331610322, 0.431267684558406, 0.485479504335672, 0.547044737264514, 0.971639374503866, 0.633815504843369, 0.177657281281427, 0.861067592864856, 0.203757208539173, 0.0928593226708472, 0.173626320436597, 0.745064198272303, 0.837384962243959, 0.892410588217899, 0.548373814672232, 0.443957264069468, 0.741993924602866, 0.914273279486224, 0.846760821063071, 0.567538046045229, 0.868789102183655, 0.386617628857493, 0.764913102146238, 0.0427200810518116, 0.787535980110988, 0.821346570504829, 0.412964291870594, 0.863511440344155, 0.594627285841852, 0.285353337414563, 0.914021850796416, 0.290276804938912, 0.396665198495612, 0.817294700304046, 0.122901547467336, 0.402618936728686, 0.49256203789264, 0.0281780988443643, 0.875322575680912, 0.584237885661423, 0.0683991103433073, 0.25067803519778, 0.356985300779343, 0.506871567340568, 0.932569323573261, 0.658679941901937, 0.894596215104684, 0.189953695284203, 0.0754130084533244, 0.387451212154701, 0.453771386295557, 0.148989676497877, 0.549559091217816, 0.949109265347943, 0.259893573820591, 0.93571306229569, 0.104800872504711, 0.835589152062312, 0.66408773441799, 0.810801792424172, 0.690951722674072, 0.991962960688397, 0.814527807058766, 0.524050284177065, 0.303657881217077, 0.1417252025567, 0.981060403166339, 0.580229231389239, 0.659985980717465, 0.626157527789474, 0.719792166957632, 0.331857268698514, 0.373319825157523, 0.60198310459964, 0.644421852426603, 0.96371345827356, 0.916483919136226, 0.349252339219674, 0.69131556013599, 0.189702027477324, 0.0264622971881181, 0.919443952850997, 0.840096372645348, 0.80028680129908, 0.963882910786197, 0.841816002968699, 0.213830050081015, 0.589360078796744, 0.450787596870214, 0.871322882827371, 0.497384141432121, 0.445205332711339, 0.249798944452778, 0.793820442166179, 0.797882409300655, 0.548264221986756, 0.85628465632908, 0.966740615200251, 0.935942951822653, 0.178680587559938, 0.368653778219596, 0.350646926090121, 0.480010354658589, 0.259779137093574, 0.588004370685667, 0.898614567006007, 0.462488604476675, 0.218185635283589, 0.314469736535102, 0.6945997599978, 0.0221551689319313, 0.868937425781041, 0.165603334084153, 0.215532686561346, 0.822113088564947, 0.274371002800763, 0.213372861267999, 0.357422668719664, 0.222921766573563, 0.249909054255113, 0.326957705197856, 0.920425582909957, 0.0215865192003548, 0.731233791215345, 0.102718993322924, 0.463496854528785, 0.880492856027558, 0.674646320287138, 0.456444133073092, 0.443644676357508, 0.560325477039441, 0.0466551757417619, 0.616055099060759, 0.816018009325489, 0.841795092448592, 0.516241285251454, 0.121707673650235, 0.462054395815358, 0.839238855056465, 0.0585555515717715, 0.609259497839957, 0.340949979610741, 0.528872431023046, 0.51486034062691, 0.0212318615522236, 0.924739596899599, 0.176980277756229, 0.840501122409478, 0.806622772943228, 0.0615717386826873, 0.731442923424765, 0.0170204245951027, 0.266738126520067, 0.789983524708077, 0.356836893828586, 0.637237603543326, 0.40373125881888, 0.778522752225399, 0.24911813181825, 0.659225017065182, 0.514489792985842, 0.358731092186645, 0.727905097883195, 0.227398766204715, 0.426877742866054, 0.517388054868206, 0.8943301809486, 0.369015514384955, 0.455049029085785, 0.283277321374044, 0.493900010362267, 0.374055189546198, 0.880992580205202, 0.568013114156201, 0.783815710339695, 0.87281924020499, 0.122833515750244, 0.766573040280491, 0.323293104767799, 0.241888492368162, 0.162471153773367, 0.262378374813125, 0.0356402245815843, 0.7577257566154, 0.945866716792807, 0.385037213563919, 0.651185262016952, 0.109772969968617, 0.143929672194645, 0.558860252611339, 0.95449004904367, 0.406574229244143, 0.803381419740617, 0.782856830628589, 0.468315722187981, 0.336180290672928, 0.053245214978233, 0.110307946568355, 0.577116705011576, 0.11026614299044, 0.888208309188485, 0.260627000359818, 0.900408041896299, 0.0284111048094928, 0.747885402757674, 0.664200517116114, 0.437765777809545, 0.962337313452736, 0.200280456105247, 0.784581347834319, 0.849413770949468, 0.892564730718732, 0.622912441380322, 0.389289381448179, 0.690327196149155, 0.928846386028454, 0.42283494467847, 0.311170934699476, 0.0520702668000013, 0.00723422458395362, 0.923357166582718, 0.961953575955704, 0.184878375846893, 0.632121226284653, 0.667110951850191, 0.176968255080283, 0.661001065280288, 0.16055632638745, 0.402860994683579, 0.422077283030376, 0.287640605121851, 0.854878806509078, 0.27881284034811, 0.909337853081524, 0.948383145965636, 0.0166410969104618, 0.787460108054802, 0.204226842382923, 0.692437045276165, 0.280433990526944, 0.106268218252808, 0.748863334069028, 0.242914773756638, 0.0480757965706289, 0.717078131157905, 0.148209916660562, 0.157302463892847, 0.588277203962207, 0.082519645569846, 0.193735569482669, 0.302687580930069, 0.625644759042189, 0.932654016185552, 0.749144951812923, 0.175621235277504, 0.126874892506748, 0.396569571923465, 0.175060101319104, 0.580286698648706, 0.17018971988, 0.0708825497422367, 0.310888579115272, 0.595160312950611, 0.380097190383822, 0.338912567356601, 0.0545874433591962, 0.329935689689592, 0.431450668023899, 0.242641343502328, 0.804629222024232, 0.881088746478781, 0.965627578087151, 0.539056585170329, 0.0477569499053061, 0.859991456381977, 0.270180185325444, 0.171894680010155, 0.32888056547381, 0.994785473681986, 0.0594475043471903, 0.298370800213888, 0.403466324554756, 0.647199931088835, 0.911356828873977, 0.42945377714932, 0.786209205631167, 0.594926910940558, 0.554257859708741, 0.255557798082009, 0.438020933419466, 0.116510102292523, 0.255302266683429, 0.260437504854053, 0.0788731940556318, 0.294838628731668, 0.482432147255167, 0.355286582605913, 0.0379172011744231, 0.62215145281516, 0.405021876329556, 0.686029544798657, 0.863577194744721, 0.700070581398904, 0.593150051310658, 0.526287148939446, 0.0406631464138627, 0.382442846894264, 0.118309584911913, 0.112204857869074, 0.654510081280023, 0.876866305945441, 0.148263417882845, 0.30803166795522, 0.151636773720384, 0.340972627978772, 0.277141680708155, 0.711694026831537, 0.416986606316641, 0.671664962312207, 0.0666286377236247, 0.170127742690966, 0.949591364711523, 0.0337314382195473, 0.075419983593747, 0.209670177893713, 0.0372481511440128, 0.410119859501719, 0.533142282161862, 0.0646388742607087, 0.200217550620437, 0.979009312810376, 0.0880640416871756, 0.545258266152814, 0.193989705992863, 0.898516889661551, 0.137698174454272, 0.236764123430476, 0.662083687260747, 0.4449109188281, 0.321947530843318, 0.182410110952333, 0.718209006823599, 0.817065063398331, 0.634123624069616, 0.390208254801109, 0.501910176128149, 0.702739205909893, 0.770107387797907, 0.265004452550784, 0.591768127866089, 0.247904427116737, 0.709953043144196, 0.555819891858846, 0.847739970311522, 0.482223408529535, 0.570594887947664, 0.567838575923815, 0.776498765451834, 0.949687656015158, 0.0698876567184925, 0.961700994288549, 0.711536508519202, 0.0259494676720351, 0.177797759650275, 0.151539009995759, 0.83689840324223, 0.751292739296332, 0.0453093505930156, 0.679821803700179, 0.152135919313878, 0.55210657347925, 0.256190184969455, 0.0122073509264737, 0.545827292604372, 0.0362222262192518, 0.130186008522287, 0.0733841492328793, 0.731199147412553, 0.580270328791812, 0.0661391078028828, 0.445170520106331, 0.0230001164600253, 0.42625815840438, 0.369656851748005, 0.677181017119437, 0.389764560153708, 0.0434793082531542, 0.925452492199838, 0.44711051415652, 0.951691653579473, 0.753615684341639, 0.500228653429076, 0.142104991478845, 0.93966357735917, 0.375828773947433, 0.0970312706194818, 0.543596994597465, 0.104115925030783, 0.612458349671215, 0.461888341931626, 0.532267089933157, 0.269966992316768, 0.945748730096966, 0.583814654033631, 0.856171891093254, 0.159431525273249, 0.0750519305001944, 0.869067372987047, 0.121931815519929, 0.105061369715258, 0.104322027415037, 0.68632849608548, 0.684196487301961, 0.315084170317277, 0.495465085376054, 0.0679924236610532, 0.243130938149989, 0.437506490619853, 0.146456979680806, 0.295873046154156, 0.152180425357074), actual = c(1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.121191354002804, 0.820131256012246, 0.773025698028505, 0.0636680608149618, 0.945937864482403, 0.900532557396218, 0.60656384495087, 0.582376500591636, 0.63616881123744, 0.52857889351435, 0.274984286166728, 0.839962652884424, 0.729040501173586, 0.370914435712621, 0.350209599127993, 0.850579360034317, 0.236072523053735, 0.0287657151930034, 0.965404144255444, 0.50565537577495, 0.0169571966398507, 0.264125578803942, 0.285837155068293, 0.479509353404865, 0.332310190424323, 0.944767536828294, 0.412483177613467, 0.603173774434254, 0.461502847960219, 0.518919355003163, 0.247216243529692, 0.746340561425313, 0.708642104174942, 0.000382420606911182, 0.637283614603803, 0.992251626681536, 0.248766761040315, 0.232408918906003, 0.637834616703913, 0.941183027112857, 0.662939365720376, 0.839954372961074, 0.495480730896816, 0.374044703319669, 0.46676546218805, 0.398645881330594, 0.362435818184167, 0.267678983509541, 0.285352592356503, 0.390014037489891, 0.162241843994707, 0.687801546882838, 0.126081730937585, 0.544015755411237, 0.656462042592466, 0.789654457941651, 0.0545338415540755, 0.534235145431012, 0.446816601790488, 0.0159425209276378, 0.343500174814835, 0.0200168157462031, 0.633148602442816, 0.0773099274374545, 0.517797402571887, 0.920342396013439, 0.338097150204703, 0.553988592233509, 0.715963247232139, 0.875436751637608, 0.726057090098038, 0.60456461738795, 0.685293920105323, 0.0146470367908478, 0.442021269584075, 0.435772309545428, 0.759147488279268, 0.21578139718622, 0.39267400209792, 0.462243297370151, 0.472040777793154, 0.241364182904363, 0.888014178257436, 0.877639832906425, 0.129783304408193, 0.214296969119459, 0.913220021873713, 0.645506735891104, 0.255426157731563, 0.830669213319197, 0.102983650518581, 0.526362414471805, 0.619921090546995, 0.176319130929187, 0.136006933171302, 0.629752807784826, 0.452721971319988, 0.423185683786869, 0.827170304488391, 0.561612982302904, 0.210193363251165, 0.0859352552797645, 0.777396853780374, 0.685387316159904, 0.925756896845996, 0.807630260940641, 0.263373618712649, 0.409156238893047, 0.145413952879608, 0.787960520014167, 0.4991863979958, 0.606943366583437, 0.459448518697172, 0.767454880988225, 0.934339116094634, 0.84473050897941, 0.215139137580991, 0.13484721374698, 0.0197513131424785, 0.698120390996337, 0.669398542493582, 0.804231449496001, 0.369638860691339, 0.575641718925908, 0.780217721359804, 0.458776701241732, 0.342167468741536, 0.34844380733557, 0.2198254710529, 0.0225933780893683, 0.691775732208043, 0.421428060391918, 0.225858493475243, 0.612071265233681, 0.990409776801243, 0.791918034432456, 0.0130674152169377, 0.963798589771613, 0.646537052234635, 0.708900143392384, 0.188801796175539, 0.0805337310303003, 0.902687831083313, 0.428523725597188, 0.723092017229646, 0.727911522844806, 0.975370053900406, 0.323145621689036, 0.376740443520248, 0.387806099141017, 0.590167665155604, 0.495996189769357, 0.372474424540997, 0.457916017156094, 0.0763549818657339, 0.889781995676458, 0.0321742419619113, 0.107280477648601, 0.266593150096014, 0.0331154593732208, 0.853530087741092, 0.734489679569378, 0.0239448614884168, 0.391049576923251, 0.114076699595898, 0.880632187006995, 0.104741991264746, 0.700342457741499, 0.643679734552279, 0.991836478468031, 0.566736238310114, 0.528420248767361, 0.271740038180724, 0.302360920468345, 0.664007309358567, 0.899373105494305, 0.170953923137859, 0.788425223669037, 0.219542416743934, 0.0777872081380337, 0.985489041777328, 0.724275218090042, 0.924368444597349, 0.335392903303728, 0.823727637063712, 0.30872166575864, 0.973839511163533, 0.346293753944337, 0.144678448326886, 0.0973479978274554, 0.226034955587238, 0.0813444836530834, 0.3248348205816, 0.423615599516779, 0.870903459377587, 0.286024479195476, 0.133188938256353, 0.171376717742532, 0.310391719918698, 0.4450190984644, 0.782150659011677, 0.786135668866336, 0.612195661524311, 0.233497101347893, 0.118386595509946, 0.735347316367552, 0.162618109257892, 0.559463962679729, 0.724708469118923, 0.23505602683872, 0.715715524740517, 0.121574015589431, 0.480720468331128, 0.811746549792588, 0.0175947528332472, 0.457108354428783, 0.872781717218459, 0.450079183094203, 0.927563693607226, 0.445924635743722, 0.150214554276317, 0.178888452937827, 0.241899957181886, 0.816931167850271, 0.926192788872868, 0.89394067064859, 0.334580505499616, 0.837795131374151, 0.0933823499362916, 0.891903295880184, 0.336640835274011, 0.595976635115221, 0.236612907610834, 0.0428265181835741, 0.366637216415256, 0.299076153198257, 0.751517867436633, 0.540223356103525, 0.57245033280924, 0.119172984501347, 0.12111559486948, 0.647838156903163, 0.414995092898607, 0.167194043286145, 0.229973666137084, 0.0165527414064854, 0.792478664778173, 0.416147246491164, 0.137791579123586, 0.304820173885673, 0.364396560238674, 0.330359725281596, 0.804430488031358, 0.501046996563673, 0.905975686851889, 0.797829643357545, 0.464782491326332, 0.647861654637381, 0.290289324941114, 0.521637263242155, 0.83149904734455, 0.100341249024495, 0.856984006473795, 0.713092073332518, 0.238103258656338, 0.460694026667625, 0.372991106938571, 0.353570340899751, 0.34919036901556, 0.264485904481262, 0.740682272939011, 0.417126327054575, 0.93319698004052, 0.0946420808322728, 0.507617083378136, 0.745390341384336, 0.961888722609729, 0.530425835167989, 0.884260697755963, 0.901732718804851, 0.448363758390769, 0.400577173568308, 0.427269470179453, 0.899847603403032, 0.146331088617444, 0.15364061226137, 0.760730826528743, 0.884378593415022, 0.323906451463699, 0.669028817443177, 0.749064846429974, 0.731257051229477, 0.450335029978305, 0.793699962552637, 0.917750927852467, 0.791846573352814, 0.255153293022886, 0.309258917113766, 0.810681942850351, 0.423636199440807, 0.930328547256067, 0.623239140259102, 0.300894979387522, 0.182865796377882, 0.216704449849203, 0.119397203903645, 0.552568964194506, 0.754814040847123, 0.4548917170614, 0.521498780930415, 0.0854400375392288, 0.463605379220098, 0.850537373451516, 0.577464552363381, 0.126719180028886, 0.28557061124593, 0.565103292930871, 0.465823726495728, 0.979797308798879, 0.687094652559608, 0.318826852133498, 0.849636507686228, 0.725227516144514, 0.671741331461817, 0.902018323540688, 0.709757149219513, 0.670706903096288, 0.336508752079681, 0.536099465563893, 0.264613440493122, 0.195595131255686, 0.746612930670381, 0.862245919182897, 0.435776767786592, 0.739192463457584, 0.169265128206462, 0.277335851918906, 0.998981047188863, 0.486176438862458, 0.559817206813022, 0.00661498587578535, 0.0937368958257139, 0.803065050160512, 0.919440344208851, 0.42102492810227, 0.700843970058486, 0.692808419698849, 0.7246605057735, 0.719005785416812, 0.696008720202371, 0.955049142474309, 0.384360111085698, 0.918907139217481, 0.591347839450464, 0.273366902256384, 0.786473701708019, 0.272870681481436, 0.687888342188671, 0.787967153824866, 0.546561791561544, 0.379295745398849, 0.258090177550912, 0.0625472429674119, 0.0277453321032226, 0.151540891733021, 0.672689920989797, 0.698512465925887, 0.354439170565456, 0.417644402943552, 0.244096543174237, 0.782153981970623, 0.611800657818094, 0.712770874844864, 0.355625098804012, 0.937263946747407, 0.933167122071609, 0.422926848521456, 0.240139890229329, 0.835353386355564, 0.260199729120359, 0.506903423229232, 0.161386760184541, 0.076601869193837, 0.0297470430377871, 0.355693113291636, 0.514979168539867, 0.344639638671651, 0.13438500254415, 0.324170573381707, 0.683039306197315, 0.13029981427826, 0.828040796332061, 0.954128982499242, 0.220239191083238, 0.965989683521912, 0.649665347766131, 0.415280659450218, 0.664990397868678, 0.064049735898152, 0.268549007829279, 0.363816239405423, 0.963908358709887, 0.121807864867151, 0.0975183569826186, 0.746485750656575, 0.0770197019446641, 0.822849649470299, 0.616522438824177, 0.525058354251087, 0.960835552308708, 0.468970155343413, 0.478926053503528, 0.012034029699862, 0.953120088903233, 0.0980712394230068, 0.756839778507128, 0.511382959783077, 0.975966026540846, 0.578093631425872, 0.351228721439838, 0.891502461628988, 0.712419027462602, 0.161058786790818, 0.508994155330583, 0.583759145345539, 0.669738958589733, 0.687045405153185, 0.867601426783949, 0.66430182219483, 0.0828949662391096, 0.354530103737488, 0.381052692886442, 0.858259873697534, 0.634058765135705, 0.976884154370055, 0.494614553637803, 0.725663950899616, 0.482763848733157, 0.126024627592415, 0.161393809597939, 0.620287484256551, 0.13536599650979, 0.432421630946919, 0.686946065630764, 0.264448006404564, 0.0846271337941289, 0.3669630989898, 0.187660786323249, 0.196607844904065, 0.665103433653712, 0.377975106937811, 0.352789320284501, 0.37673553568311, 0.302581144729629, 0.185588057152927, 0.00861090933904052, 0.693774171872064, 0.910567098995671, 0.263572215102613, 0.809791217325255, 0.361887383274734, 0.200854879338294, 0.375781887210906, 0.0144633124582469, 0.00971146975643933, 0.0011172229424119, 0.990935907466337, 0.905863877385855, 0.921737437136471, 0.825439155567437, 0.451621573651209, 0.708591204602271, 0.139061665860936, 0.809131562476978, 0.950669816927984, 0.430104164406657, 0.267841712804511, 0.742614340269938, 0.165722080273554, 0.439579728757963, 0.045111934421584, 0.0624599307775497, 0.189184522256255, 0.89072443684563, 0.903473081300035, 0.998555902391672, 0.713943479117006, 0.680620609782636, 0.614467471605167, 0.360596258658916, 0.536184039432555, 0.194547132588923, 0.317299195099622, 0.0178075691219419, 0.948348503559828, 0.398962036939338, 0.241111156763509, 0.393383968621492, 0.135017446707934, 0.51107216742821, 0.983105866704136, 0.705370904412121, 0.958885811967775, 0.457275861874223, 0.705589016433805, 0.400327499723062, 0.512462677666917, 0.815371653530747, 0.674107776023448, 0.328288460616022, 0.356046231696382, 0.0719508386682719, 0.292623973917216, 0.189461536705494, 0.43552332255058, 0.692229049513116, 0.904385315487161, 0.182303840992972, 0.73400383698754, 0.96045626164414, 0.260129011934623, 0.216673995833844, 0.307481834664941, 0.453500772127882, 0.907135885208845, 0.189720394322649, 0.285636044573039, 0.454813356976956, 0.924871810944751, 0.731300421757624, 0.99843606329523, 0.36908484948799, 0.512463500490412, 0.318785607349128, 0.339422651333734, 0.668378184083849, 0.0522843031212687, 0.115263623185456, 0.0051210462115705, 0.0145909897983074, 0.707803047262132, 0.907401652308181, 0.155467019882053, 0.951684875413775, 0.413035499164835, 0.303113889647648, 0.485771861160174, 0.708095788024366, 0.992752053774893, 0.288669745437801, 0.0692982752807438, 0.00352690648287535, 0.568335723830387, 0.64842953812331, 0.096666157245636, 0.441060768440366, 0.297732590464875, 0.372808993328363, 0.560483899433166, 0.60663118143566, 0.848329218337312, 0.522318379953504, 0.389189674286172, 0.803202393464744, 0.721422550734133, 0.985113541828468, 0.614068342605606, 0.900442081969231, 0.0731347834225744, 0.0265320595353842, 0.565854453947395, 0.97029332537204, 0.55279625300318, 0.831372078042477, 0.157967939740047, 0.520176561083645, 0.64876364544034, 0.417676161276177, 0.744265257148072, 0.344132523983717, 0.639762627426535, 0.759746036725119, 0.85480456892401, 0.636991246370599, 0.0835894758347422, 0.861066384706646, 0.0834769764915109, 0.105333912884817, 0.335110600572079, 0.654915245482698, 0.577653096057475, 0.399729291908443, 0.0173243782483041, 0.23514884733595, 0.488617609487846, 0.0183238848112524, 0.842492436990142, 0.426108250394464, 0.63480879063718, 0.5545091512613, 0.962972914800048, 0.760417833225802, 0.389211213914678, 0.469838383840397, 0.00205864710733294, 0.305202728137374, 0.947859764797613, 0.284266388276592, 0.980376892257482, 0.640781265450642, 0.693831199780107, 0.429224183550104, 0.0660306510981172, 0.76617063744925, 0.906185632338747, 0.753204726846889, 0.731002515181899, 0.418079154798761, 0.770815652096644, 0.624766878085211, 0.831030725268647, 0.031164510641247, 0.715867628576234, 0.0412917616777122, 0.197474331595004, 0.888764190254733, 0.480850903317332, 0.11448887269944, 0.557526748394594, 0.16271925275214, 0.306138137122616, 0.688688572030514, 0.119574287440628, 0.640691533917561, 0.546645486494526, 0.872978824656457, 0.752933578332886, 0.296888181008399, 0.128012711647898, 0.926236530998722, 0.266623053699732, 0.132529909256846, 0.546975038014352, 0.304502197308466, 0.0337308205198497, 0.0476977950893342, 0.350888328859583, 0.193396401591599, 0.313684229971841, 0.886404657270759, 0.436834907159209, 0.516340981703252, 0.708762881113216, 0.566478107590228, 0.868478094460443, 0.305597309721634, 0.91507520666346, 0.982438307488337, 0.569605152588338, 0.0151422913186252, 0.823956296779215, 0.0024405971635133, 0.176779667846859, 0.679628258105367, 0.0656105466187, 0.454912239452824, 0.0749848193954676, 0.0363571289926767, 0.824814349180087, 0.953018415952101, 0.926430045161396, 0.231865913141519, 0.712791074765846, 0.254274077014998, 0.673504400998354, 0.654429107904434, 0.918133534723893, 0.0454632425680757, 0.64573148614727, 0.970638589235023, 0.657940661767498, 0.0238792051095515, 0.825563699007034, 0.836294987238944, 0.868656464386731, 0.43523660977371, 0.645313285291195, 0.0226292237639427, 0.0387158538214862, 0.846807015826926, 0.0509231612086296, 0.166835140436888, 0.659880070481449, 0.777625131886452, 0.560724401148036, 0.53076493437402, 0.0292208176106215, 0.651772820390761, 0.834545638179407, 0.502970546251163, 0.42328505218029, 0.371622524457052, 0.00426665763370693, 0.857504269108176, 0.25654680468142, 0.789905309909955, 0.666222564177588, 0.382418847410008, 0.666229740250856, 0.214025810360909, 0.438664364395663, 0.616064889822155, 0.407779759960249, 0.564280040794984, 0.40455457335338, 0.603309670928866, 0.188641466666013, 0.253499347949401, 0.397085355827585, 0.821420247899368, 0.408979444298893, 0.872884290991351, 0.366620550630614, 0.250002224231139, 0.304464464308694, 0.968055358389392, 0.675731349270791, 0.4903906180989, 0.47728890622966, 0.276956908870488, 0.811038173502311, 0.962120260111988, 0.66097024967894, 0.344861160265282, 0.83195382845588, 0.727714616106823, 0.104722973424941, 0.183298902586102, 0.79891298012808, 0.027250207029283, 0.229968699859455, 0.852449353085831, 0.116104824468493, 0.479730397695675, 0.325177826918662, 0.700928409118205, 0.520243662176654, 0.645262336358428, 0.413851663470268, 0.0224581407383084, 0.688280375441536, 0.156655150698498, 0.51051390520297, 0.775885947747156, 0.0497170318849385, 0.56985977338627, 0.664731195895001, 0.14716147701256, 0.426383510930464, 0.554166654823348, 0.0385742033831775, 0.731585357803851, 0.501092238351703, 0.75570821785368, 0.611406147480011, 0.079736594343558, 0.558824996929616, 0.713634991087019, 0.477269171504304, 0.956642253790051, 0.72862455714494, 0.442165852058679, 0.190085727488622, 0.0151893680449575, 0.801288582617417, 0.618512877030298, 0.00740615767426789, 0.75516743469052, 0.208744253497571, 0.832216661889106, 0.112034918973222, 0.132126061944291, 0.583437647903338, 0.0617478520143777, 0.341036544181406, 0.931123893940821, 0.00441295281052589, 0.650297866202891, 0.911423484329134, 0.366066077025607, 0.286503884010017, 0.889300321694463, 0.217855305876583, 0.568608587374911, 0.819570162799209, 0.346717149484903, 0.684672572184354, 0.190371710341424, 0.257739912718534, 0.97371239005588, 0.894436567090452, 0.00962219922803342, 0.716542424168438, 0.966164621291682, 0.388234766433015, 0.384998820722103, 0.241839520633221, 0.635487376945093, 0.973649990512058, 0.209534492809325, 0.345215615816414, 0.0888365311548114, 0.64908993570134, 0.15131273213774, 0.273640861501917, 0.790460636373609, 0.67320044641383, 0.980604207608849, 0.104457858251408, 0.723745589144528, 0.792768886312842, 0.233407363528386, 0.251235859002918, 0.919311531819403, 0.816741451388225, 0.199676880147308, 0.171388109214604, 0.788893623743206, 0.604305719025433, 0.302498625824228, 0.528565464774147, 0.334912218851969, 0.691153792897239, 0.335502983536571, 0.176022351486608, 0.123281450243667, 0.504492959706113, 0.705043805064633, 0.944350062403828, 0.689927339088172, 0.204869041917846, 0.494061641627923, 0.184701038524508, 0.817971856333315, 0.706153575796634, 0.922947789309546, 0.499415511032566, 0.818576822290197, 0.337648158427328, 0.844123841030523, 0.248196405824274, 0.7972678525839, 0.264929220546037, 0.0984420923050493, 0.927828608779237, 0.249763773754239, 0.578960442915559, 0.279832290718332, 0.740711129503325, 0.306131039047614, 0.420791587093845, 0.510449904948473, 0.484037282178178, 0.350102143594995, 0.282135978341103, 0.53794557438232, 0.652869523735717, 0.871353311697021, 0.312263406580314, 0.552750078961253, 0.385386033682153, 0.868759607896209, 0.183970261132345, 0.299034407362342, 0.423311229096726, 0.955971753923222, 0.292165219318122, 0.492836651857942, 0.226916674291715, 0.491359986830503, 0.981794329360127, 0.701120231067762, 0.302714047487825, 0.863873402355239, 0.00879155704751611, 0.185326071223244, 0.62392903259024, 0.156939880223945, 0.623796122148633, 0.45613284711726, 0.0696764718741179, 0.3909082908649, 0.258942366577685, 0.468289507785812, 0.874575267545879, 0.529651142889634, 0.292280764319003, 0.877999919466674, 0.702331458218396, 0.991823844378814, 0.110418013529852, 0.124704682966694, 0.716186590958387, 0.240513149648905, 0.777888610493392, 0.682832452235743, 0.0433035704772919, 0.423533624503762, 0.683981875190511, 0.173706435598433, 0.29690523352474, 0.406183829763904, 0.25022635422647, 0.516484500607476, 0.210870372597128, 0.439300311962143, 0.019302295288071, 0.547939385054633, 0.427121065091342, 0.376011717831716, 0.0924288553651422, 0.0586172486655414, 0.135165626415983, 0.834737193072215, 0.573852749075741, 0.493117341771722, 0.249676872976124, 0.182533679530025, 0.40262617636472, 0.139760201331228, 0.372332191327587, 0.785486505366862, 0.519215799402446, 0.499056506436318, 0.551120816031471, 0.950763259548694, 0.47991503495723, 0.892817242769524, 0.489509493811056, 0.328198393806815, 0.871802878798917, 0.157004881417379, 0.896162847522646, 0.9747061021626, 0.844983955612406, 0.910844983067364, 0.34670454566367, 0.961241822456941, 0.209364566719159, 0.907265706220642, 0.490908426465467, 0.54783217725344, 0.349144812207669, 0.135747471824288, 0.115743251750246, 0.288072820054367, 0.584937214385718, 0.812237231759354, 0.476844450226054, 0.863172588171437, 0.275850178673863, 0.0110326278954744, 0.0880273166112602, 0.688471793662757, 0.746591142844409, 0.996542427921668, 0.755852636182681, 0.167371237650514, 0.932352898875251, 0.346861559664831, 0.799732656683773, 0.0306277857162058, 0.335375407012179, 0.233958726283163, 0.38212152919732, 0.40424917708151, 0.618472381494939, 0.0556291162502021, 0.407193405553699, 0.727201235713437, 0.402076800586656, 0.778047658968717, 0.575734915211797, 0.256792027037591, 0.793647547485307, 0.560045188525692, 0.466418629512191, 0.0762160182930529, 0.367293404880911, 0.70518002496101, 0.814570032060146, 0.676414167508483, 0.0972384072374552, 0.893685087095946, 0.0338346576318145, 0.393980292137712, 0.0902431088034064, 0.385111462324858, 0.387614127947018, 0.966520932503045), actual = c(0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"