Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.5868627        0.5396079 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5396079        0.5396079        0.5012763 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual}
# as arguments. They can also take \code{nmb} (named vector containing NMB
# with values for TP, FP, TN, FN).
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.695947024738416, 0.638544905697927, 0.166246100794524, 0.166356671368703, 0.836421615676954, 0.338108968921006, 0.580315183149651, 0.0819806621875614, 0.231972160981968, 0.267207693308592, 0.235898409737274, 0.218234974890947, 0.62682082480751, 0.838858153438196, 0.739400176331401, 0.74876586953178, 0.0227007993962616, 0.617593909613788, 0.329933246364817, 0.709026054944843, 0.369533355114982, 0.201524604577571, 0.0384939522482455, 0.596319221192971, 0.996309542795643, 0.682288079056889, 0.65395112358965, 0.464223376940936, 0.495828591287136, 0.538800973678008, 0.224926376016811, 0.646009182557464, 0.914527154294774, 0.110720889875665, 0.767671010456979, 0.266904493561015, 0.804816305404529, 0.112858731532469, 0.217040824703872, 0.544388002483174, 0.369355834554881, 0.156056586420164, 0.374222149839625, 0.944022873183712, 0.0994732193648815, 0.654024597490206, 0.504190438892692, 0.797763420501724, 0.0448601313401014, 0.571331018349156, 0.451105065178126, 0.751947147306055, 0.322877082042396, 0.664969855453819, 0.685817567398772, 0.935085454955697, 0.486988609423861, 0.758170212851837, 0.477273285156116, 0.794317069696262, 0.204426556825638, 0.181145865703002, 0.936077717458829, 0.528192438883707, 0.856381841236725, 0.0566448229365051, 0.845565533265471, 0.762532678199932, 0.542937808204442, 0.727571775438264, 0.248263611691073, 0.545182258822024, 0.674110793042928, 0.60714633250609, 0.702017523813993, 0.464339795988053, 0.894936883822083, 0.756412795744836, 0.435866161016747, 0.807802614057437, 0.754609605995938, 0.299016101518646, 0.0300672226585448, 0.083908898755908, 0.456852925708517, 0.672461721813306, 0.505391530692577, 0.837512217694893, 0.973781984066591, 0.931201280094683, 0.635847464436665, 0.387588739860803, 0.914236797485501, 0.521840553963557, 0.707145872991532, 0.56772822397761, 0.30426083272323, 0.0246813322883099, 0.308542129583657, 0.205371217802167, 0.820895007578656, 0.0209215581417084, 0.0766607597470284, 0.00612826459109783, 0.0175431575626135, 0.378038431983441, 0.63102935696952, 0.585702976677567, 0.181746469810605, 0.605030091013759, 0.38410929730162, 0.384033282287419, 0.0490873528178781, 0.0984943094663322, 0.0984050410334021, 0.464285087771714, 0.536910936469212, 0.845954573014751, 0.474015835206956, 0.543282182887197, 0.421450678491965, 0.497160400031134, 0.270007949555293, 0.462889006361365, 0.413463238393888, 0.179333773208782, 0.793746907031164, 0.207126542693004, 0.973297691438347, 0.914520437596366, 0.241952452110127, 0.831151582067832, 0.856737327063456, 0.673130702460185, 0.928424491314217, 0.580089963739738, 0.4634830635041, 0.816433937288821, 0.698692354140803, 0.815571951214224, 0.144322798354551, 0.243293511448428, 0.989036249462515, 0.0171809194143862, 0.605859311064705, 0.107779669109732, 0.306368140038103, 0.842345444485545, 0.927306529600173, 0.915202734759077, 0.153924659593031, 0.487320324638858, 0.993255942361429, 0.848190706688911, 0.70597721892409, 0.723323851358145, 0.998198436340317, 0.670554882846773, 0.00408750819042325, 0.525387792382389, 0.546660756226629, 0.232799012213945, 0.429205594584346, 0.614925445523113, 0.454035954782739, 0.722148171858862, 0.800333281280473, 0.439826916670427, 0.707307592500001, 0.689099189359695, 0.171551254577935, 0.895590892992914, 0.654832661617547, 0.480993523728102, 0.853587637422606, 0.505865153390914, 0.944631986552849, 0.0140414459165186, 0.6281609502621, 0.915011280449107, 0.492327726446092, 0.416754597099498, 0.246772279497236, 0.166620512260124, 0.189634513109922, 0.268099032575265, 0.530054706148803, 0.769379853736609, 0.916088512632996, 0.280898088123649, 0.305061266291887, 0.547321433434263, 0.382277824450284, 0.532234118785709, 0.139794751768932, 0.634578495053574, 0.529221949633211, 0.314721108181402, 0.149690860882401, 0.219344338867813, 0.397300153737888, 0.797992372652516, 0.863716553198174, 0.97958111600019, 0.720297301188111, 0.64803563291207, 0.218274738406762, 0.145686936331913, 0.685285271145403, 0.736511179478839, 0.584620503941551, 0.401821631705388, 0.0980030877981335, 0.492698184214532, 0.318208954297006, 0.229139937320724, 0.260953351389617, 0.506395671749488, 0.60761310486123, 0.430769867729396, 0.0423628783319145, 0.698713595978916, 0.2457181410864, 0.287979552289471, 0.956168965902179, 0.778244382236153, 0.413614546181634, 0.99287372129038, 0.00793298147618771, 0.595213214401156, 0.69882228388451, 0.142283900873736, 0.803375923773274, 0.566275957971811, 0.190883615985513, 0.38175390008837, 0.569536134134978, 0.843515661079437, 0.0758115409407765, 0.0428582420572639, 0.463608837686479, 0.229040232254192, 0.768315665889531, 0.832185465609655, 0.820184841519222, 0.562425010371953, 0.556055163498968, 0.677049072226509, 0.39775602100417, 0.380484852474183, 0.15703886654228, 0.282446483615786, 0.976587921148166, 0.477105878526345, 0.319942873436958, 0.237352048279718, 0.0287300725467503, 0.823716765502468, 0.747307104989886, 0.830457178875804, 0.361327884020284, 0.676961605902761, 0.161201362963766, 0.733088323147967, 0.905620228499174, 0.142367902677506, 0.470551965059713, 0.456470270641148, 0.466486212331802, 0.742682936834171, 0.715258392039686, 0.989124142332003, 0.0251749868039042, 0.0409716118592769, 0.309687686618418, 0.42462049680762, 0.497348241042346, 0.446036969544366, 0.829572910442948, 0.271309578791261, 0.712092098314315, 0.495334841776639, 0.0721097115892917, 0.167583921924233, 0.338087239302695, 0.716827417956665, 0.764014380984008, 0.379225468495861, 0.26393931498751, 0.789682662812993, 0.684625918976963, 0.950064914068207, 0.58828710950911, 0.0123647281434387, 0.352312622359022, 0.880397137254477, 0.059770338004455, 0.298260989133269, 0.961474074050784, 0.849874066421762, 0.549753546249121, 0.479602687060833, 0.305529868695885, 0.355217440752313, 0.117292821407318, 0.348095831694081, 0.43676500627771, 0.98562144045718, 0.770273309899494, 0.800317907705903, 0.317187300184742, 0.528407371370122, 0.822428079554811, 0.940786702092737, 0.746637673582882, 0.230960212647915, 0.555811248254031, 0.767242441186681, 0.403965645935386, 0.95897804829292, 0.918388670310378, 0.160101455170661, 0.456181040266529, 0.0408878109883517, 0.980171508388594, 0.507362595526502, 0.134126977995038, 0.261680233757943, 0.0430500076618046, 0.631327910115942, 0.367902484256774, 0.862432949477807, 0.95590648567304, 0.885350453201681, 0.370666780974716, 0.307296330109239, 0.440239539835602, 0.72352739633061, 0.291720303008333, 0.471463285386562, 0.0554327864665538, 0.657296415418386, 0.370742227882147, 0.543642149772495, 0.846191091230139, 0.502756953937933, 0.106267714872956, 0.118519233539701, 0.204704201314598, 0.0810580377001315, 0.480919569265097, 0.91707257879898, 0.106866408372298, 0.805329149821773, 0.229159825015813, 0.447917253244668, 0.162925042677671, 0.0787981844041497, 0.718134015565738, 0.0757203598041087, 0.884998706635088, 0.275042835855857, 0.646095905918628, 0.207675523590297, 0.775679079815745, 0.75071760523133, 0.33015054021962, 0.408456482924521, 0.412591596134007, 0.443024984328076, 0.505584685364738, 0.549316285876557, 0.605417157756165, 0.809943958651274, 0.662525810766965, 0.8346289137844, 0.618989122565836, 0.13946322328411, 0.843394652940333, 0.440254850080237, 0.155597284669057, 0.624985836446285, 0.833083694102243, 0.156023045768961, 0.78949208999984, 0.529084604466334, 0.47034848597832, 0.290677826385945, 0.452240078477189, 0.427298714639619, 0.127213294617832, 0.058676395798102, 0.515348511282355, 0.764398673549294, 0.457230390748009, 0.362851933808997, 0.967631104402244, 0.793445134535432, 0.329433051170781, 0.493970220210031, 0.165145149454474, 0.929875010857359, 0.984663547948003, 0.207335154991597, 0.878316728165373, 0.989951350493357, 0.351891315774992, 0.920469685690477, 0.708808958064765, 0.896880987333134, 0.511106740450487, 0.471842255908996, 0.316157263237983, 0.0236222720704973, 0.183240058133379, 0.0153458062559366, 0.604193100705743, 0.730599491624162, 0.2719576661475, 0.812352738343179, 0.781992502976209, 0.928734795656055, 0.647861753124744, 0.340070024365559, 0.62933399179019, 0.556062438990921, 0.458934772294015, 0.364151573972777, 0.626024252967909, 0.0805580841843039, 0.1480035581626, 0.480717282276601, 0.798071507364511, 0.486085254931822, 0.381616437807679, 0.921164341969416, 0.131510565057397, 0.184255935717374, 0.831403153482825, 0.307355597848073, 0.54535706830211, 0.218446189071983, 0.430206642253324, 0.0577668002806604, 0.409425171092153, 0.352333537535742, 0.925241473596543, 0.399021889781579, 0.0198045163415372, 0.549842923879623, 0.299092043656856, 0.72865796694532, 0.752201529452577, 0.184453884139657, 0.32416988350451, 0.00249894009903073, 0.243380293017253, 0.198337017092854, 0.265565450536087, 0.0972275491803885, 0.532680474920198, 0.0114120696671307, 0.0516421019565314, 0.143364559626207, 0.558014536043629, 0.762690798379481, 0.914233240066096, 0.961729872273281, 0.00640913820825517, 0.701419888064265, 0.929910792969167, 0.375601107021794, 0.550096317660064, 0.0298396155703813, 0.537683086236939, 0.123426209902391, 0.357415585080162, 0.209467231994495, 0.0755430199205875, 0.422988961217925, 0.240997679065913, 0.61216937401332, 0.532251334981993, 0.254323727684096, 0.807482351316139, 0.49796582525596, 0.194818737916648, 0.489026899449527, 0.949688460445032, 0.481575041310862, 0.516716262092814, 0.410226620733738, 0.457733749644831, 0.441979181952775, 0.232305252691731, 0.325263281119987, 0.262319224886596, 0.91876851185225, 0.838166889036074, 0.782618715427816, 0.547448672121391, 0.75308801792562, 0.673906334443018, 0.089611642062664, 0.211482918355614, 0.795574937947094, 0.238001845311373, 0.83563199127093, 0.17221604171209, 0.913833134109154, 0.99327609851025, 0.334234219277278, 0.372564826626331, 0.0142329062800854, 0.852581390412524, 0.822408305481076, 0.885075073223561, 0.831571752438322, 0.612272103782743, 0.0188857277389616, 0.285349293844774, 0.770836960989982, 0.863078635651618, 0.824051144998521, 0.97484819800593, 0.158460781211033, 0.917519418522716, 0.913167170016095, 0.439679205417633, 0.798196032876149, 0.132175847887993, 0.706337188836187, 0.172018836485222, 0.671282052993774, 0.405947651481256, 0.285000432515517, 0.983623035717756, 0.716091762762517, 0.494928856147453, 0.05898208450526, 0.399415501160547, 0.649818367091939, 0.675793197005987, 0.911030894611031, 0.988488633884117, 0.646379170240834, 0.546513319248334, 0.0383155245799571, 0.750939219491556, 0.387568109901622, 0.242264993255958, 0.244793503312394, 0.770065165823326, 0.255112466635182, 0.131771513726562, 0.540075870929286, 0.787118242122233, 0.707700881641358, 0.558326985687017, 0.503113685641438, 0.0738318958319724, 0.409928368171677, 0.715897613903508, 0.453215231420472, 0.245422079693526, 0.468097315635532, 0.485700543271378, 0.102594948140904, 0.586303029675037, 0.32962356461212, 0.534056517761201, 0.0730017456226051, 0.956378723494709, 0.469873609486967, 0.434607671806589, 0.447152371518314, 0.771955192787573, 0.259653853718191, 0.736471622716635, 0.933780299965292, 0.435226685134694, 0.0511066112667322, 0.343925748486072, 0.947605309542269, 0.0212011961266398, 0.838352717459202, 0.450783731881529, 0.8159681011457, 0.0127424779348075, 0.0879759027156979, 0.465253862552345, 0.414393156766891, 0.392841714899987, 0.858753006206825, 0.384976202622056, 0.329375576926395, 0.244767927797511, 0.014895785599947, 0.582263628486544, 0.39137042616494, 0.419492071727291, 0.974465058650821, 0.268669557292014, 0.645350554957986, 0.616807717364281, 0.388732786290348, 0.699589332332835, 0.269870201824233, 0.0658607166260481, 0.974164021201432, 0.170918641146272, 0.325906740967184, 0.328008668031543, 0.23102019354701, 0.157746027456596, 0.848849282134324, 0.508529245853424, 0.0649423035793006, 0.748702588025481, 0.557465084595606, 0.91421760013327, 0.197889211587608, 0.743298397399485, 0.779680119128898, 0.57762891612947, 0.473034508759156, 0.0777549305930734, 0.939932235283777, 0.738824896514416, 0.527861895505339, 0.173029289580882, 0.0895911045372486, 0.728443830041215, 0.714692808687687, 0.504672632319853, 0.889117949875072, 0.399963383097202, 0.711442710598931, 0.976909541524947, 0.135303468210623, 0.290144434198737, 0.893256154842675, 0.866421160055324, 0.593196962727234, 0.99819471151568, 0.999408514471725, 0.65023073926568, 0.269728585844859, 0.908921404741704, 0.922133292304352, 0.848522152751684, 0.157114743255079, 0.75536346109584, 0.51213378412649, 0.11525445128791, 0.602731367805973, 0.929293055087328, 0.365877751028165, 0.29010044480674, 0.417370439041406, 0.223025813465938, 0.476697897538543, 0.547952003777027, 0.162029549945146, 0.387279933318496, 0.322855147765949, 0.616365268360823, 0.582298519322649, 0.693744238000363, 0.325946675147861, 0.183653756743297, 0.262239611707628, 0.0720739958342165, 0.74977142480202, 0.670225297566503, 0.656323041766882, 0.741499824682251, 0.308922095689923, 0.429927916033193, 0.569809007691219, 0.601753136143088, 0.374439362902194, 0.333766168681905, 0.428925490938127, 0.0923628229647875, 0.581970937084407, 0.500069789355621, 0.265886621316895, 0.0717375727836043, 0.81861079367809, 0.343072764808312, 0.47699322947301, 0.723057502415031, 0.274222607258707, 0.218557152198628, 0.299221417400986, 0.475297005381435, 0.443412956083193, 0.247251058230177, 0.531899891328067, 0.172922867815942, 0.304970176192001, 0.746726716635749, 0.304463780252263, 0.676427026744932, 0.489815074717626, 0.568730765255168, 0.650791574502364, 0.548082838766277, 0.0774716222658753, 0.650830894708633, 0.148635075194761, 0.939024267485365, 0.0104545878712088, 0.830555491382256, 0.723152694525197, 0.276273020077497, 0.372060086112469, 0.132161017972976, 0.596667184960097, 0.171261358540505, 0.991879012202844, 0.699479831615463, 0.845423787366599, 0.863175069447607, 0.779055233346298, 0.367703990312293, 0.872595290886238, 0.359027517493814, 0.958758029853925, 0.707008345285431, 0.0366086047142744, 0.20925288181752, 0.268126522190869, 0.692967109382153, 0.688357199775055, 0.240258648060262, 0.937480985885486, 0.0371095084119588, 0.559665178181604, 0.833191835787147, 0.75582144013606, 0.821048527956009, 0.792052568402141, 0.791425757808611, 0.409382563317195, 0.146955584874377, 0.167431962909177, 0.410585473524407, 0.633397359168157, 0.26593878865242, 0.035905854543671, 0.377122921869159, 0.876972977304831, 0.651650160551071, 0.558866984210908, 0.581887974170968, 0.242257452569902, 0.327385324519128, 0.153622245881706, 0.241319159977138, 0.647060685558245, 0.967733949190006, 0.0908346916548908, 0.935330527368933, 0.870568648446351, 0.361002267105505, 0.433577860705554, 0.0383159145712852, 0.335958558833227, 0.746022577164695, 0.140285161789507, 0.409745298326015, 0.845942800166085, 0.691925306571648, 0.75089241634123, 0.401578496908769, 0.884383185766637, 0.277831570478156, 0.446541121229529, 0.384296867065132, 0.576267900178209, 0.767502724658698, 0.373532727127895, 0.584851976716891, 0.52439680788666, 0.95842176489532, 0.917976211756468, 0.647884410107508, 0.173904709750786, 0.554173659067601, 0.29143900773488, 0.370624273782596, 0.743815341265872, 0.975863723317161, 0.682274639839306, 0.930116406176239, 0.705681163584813, 0.931698829401284, 0.615410696947947, 0.121894806623459, 0.31859844038263, 0.928465317003429, 0.672893893439323, 0.979478617664427, 0.984747772105038, 0.170072739478201, 0.688588473945856, 0.0869698710739613, 0.67482116445899, 0.670326254563406, 0.403338162694126, 0.247391275130212, 0.351168170804158, 0.656662618042901, 0.0144379641860723, 0.958334024762735, 0.251385119510815, 0.787025787169114, 0.383716656127945, 0.0603212784044445, 0.879453911678866, 0.221388125093654, 0.284090496832505, 0.392395051661879, 0.976095267571509, 0.220343328779563, 0.897475898265839, 0.329585573635995, 0.320151879452169, 0.629521541064605, 0.633097370387986, 0.826916540507227, 0.629504611715674, 0.790962311672047, 0.764777713222429, 0.187295076204464, 0.59720268798992, 0.275007963413373, 0.705926646245643, 0.00918503548018634, 0.274301283992827, 0.0396073961164802, 0.700158375082538, 0.729349147295579, 0.825836989097297, 0.976153990719467, 0.615915906615555, 0.613766566384584, 0.0554604483768344, 0.227431062608957, 0.438490112079307, 0.84860011911951, 0.218692951835692, 0.337614775868133, 0.575890911044553, 0.811819516122341, 0.378906231140718, 0.638515120837837, 0.202009509783238, 0.993639040039852, 0.807711235946044, 0.71116321743466, 0.358957568416372, 0.699488352751359, 0.155034997267649, 0.526201292872429, 0.530439038528129, 0.00883659906685352, 0.946759782265872, 0.642883742460981, 0.285424935398623, 0.115564807550982, 0.398228945443407, 0.86165960249491, 0.0581510402262211, 0.806659518973902, 0.568428140832111, 0.157832931727171, 0.169150697998703, 0.113532960880548, 0.786302439169958, 0.380825056228787, 0.174012047704309, 0.539147765841335, 0.36595224076882, 0.549993731081486, 0.351286286721006, 0.306306555867195, 0.630252352450043, 0.364269536919892, 0.524340229807422, 0.92016448196955, 0.178815877763554, 0.661178799811751, 0.260660331696272, 0.386810235213488, 0.210799572989345, 0.959639403270558, 0.0822776039130986, 0.868098011240363, 0.890599458245561, 0.898779863258824, 0.878062666859478, 0.0328871244564652, 0.557142419042066, 0.596158332191408, 0.698192327748984, 0.508407094050199, 0.445473898435012, 0.0306213342119008, 0.203336023492739, 0.036187257617712, 0.886749297380447, 0.235973334871233, 0.249836222501472, 0.090328756486997, 0.471148837124929, 0.461441182298586, 0.686519818613306, 0.372021134477109, 0.253073920030147, 0.831871695350856, 0.636710254941136, 0.232245781458914, 0.785442768828943, 0.165693403920159, 0.182981307152659, 0.151777054648846, 0.843555197119713, 0.766761634498835, 0.0820184713229537, 0.600341177545488, 0.227316963486373, 0.382068181410432, 0.220305456314236, 0.864333242410794, 0.794704442145303, 0.529887168901041, 0.235689497552812, 0.61041957186535, 0.734998888103291, 0.201592106372118, 0.273002541391179, 0.187250917078927, 0.00701047480106354, 0.665582751389593, 0.525100870523602, 0.911331243347377, 0.506184769328684, 0.896524650044739, 0.0177784266415983, 0.551410739077255, 0.250846010865644, 0.335535819642246, 0.169037848711014, 0.964361614082009, 0.936783149139956, 0.888304031919688, 0.0371545539237559, 0.950869095511734, 0.876535264775157, 0.292059912579134, 0.598178219283, 0.858868198236451, 0.819155951263383, 0.337260506581515, 0.740445561241359, 0.471972836647183, 0.394758229143918, 0.43525752495043, 0.585065572522581, 0.902858898276463, 0.156683231471106, 0.165567015297711, 0.811968987574801, 0.626138334162533, 0.778750046854839, 0.50917886197567, 0.553548818686977, 0.212080537108704, 0.775217064889148, 0.455872832331806, 0.350003243191168, 0.909422390628606, 0.356129652587697, 0.731504541821778, 0.678050555055961, 0.883698862977326, 0.425360399996862, 0.802970082499087, 0.90170155162923, 0.00688010570593178, 0.729221083922312, 0.653291549999267, 0.2205908105243, 0.116196398623288, 0.336063145659864, 0.161338160512969, 0.443993594264612), actual = c(1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and
# inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c(
    "fx_roc01",
    "fx_sum_sens_spec",
    "youden",
    "all",
    "none"
  ),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.963558899238706, 0.217918073758483, 0.55497309146449, 0.105430947383866, 0.572407636325806, 0.492732285987586, 0.252874702680856, 0.362072848482057, 0.444331178674474, 0.71785701951012, 0.971041848417372, 0.634778338950127, 0.296708274865523, 0.0976328179240227, 0.327567976666614, 0.291858247714117, 0.650045842863619, 0.0958573031239212, 0.007640729425475, 0.844094288302585, 0.613198146922514, 0.554992722580209, 0.0356887262314558, 0.894826266448945, 0.555237338878214, 0.951055958634242, 0.149959609843791, 0.927032121224329, 0.0186771845910698, 0.800005492055789, 0.856191885890439, 0.0546480098273605, 0.248094607610255, 0.430788430152461, 0.072154792258516, 0.311230533756316, 0.0689455517567694, 0.397150988224894, 0.289037134731188, 0.814995288150385, 0.0576347929891199, 0.674870854942128, 0.134772980120033, 0.188537450507283, 0.103169368812814, 0.37997301784344, 0.228518371703103, 0.685798361664638, 0.624801865778863, 0.841084684478119, 0.929935622494668, 0.28818705980666, 0.800919494591653, 0.618537327973172, 0.474810715764761, 0.633028003852814, 0.775561886839569, 0.898624157300219, 0.233800574671477, 0.0526632289402187, 0.553814728045836, 0.127673674374819, 0.537159320898354, 0.182085362495854, 0.151462157722563, 0.25235147960484, 0.12978192884475, 0.550476470263675, 0.214997619157657, 0.88524806802161, 0.39763896795921, 0.426313840318471, 0.686957881087437, 0.941502519184723, 0.829359378200024, 0.846451890654862, 0.435726104304194, 0.207900426350534, 0.166896186070517, 0.165644543711096, 0.579867180902511, 0.148981752339751, 0.893403866095468, 0.160388176562265, 0.854464595438913, 0.670577390119433, 0.417891069548205, 0.865780304418877, 0.836924003204331, 0.783499369863421, 0.912829317850992, 0.103176485747099, 0.890294911339879, 0.0645675894338638, 0.395849816733971, 0.0319582971278578, 0.251641527982429, 0.394833560567349, 0.837824018439278, 0.421736351214349, 0.263790184166282, 0.95201809075661, 0.451125625520945, 0.226765839150175, 0.706337244715542, 0.0653855046257377, 0.570823105284944, 0.710517989704385, 0.634570189984515, 0.890494748251513, 0.923213381785899, 0.907852546777576, 0.858882564119995, 0.996708562597632, 0.650136808399111, 0.820703792851418, 0.486070006387308, 0.924032860202715, 0.952952007763088, 0.355517806950957, 0.79091574321501, 0.937520360574126, 0.234176304889843, 0.832863437943161, 0.466824295464903, 0.862239162437618, 0.631526894168928, 0.432273343671113, 0.876800365978852, 0.206575107527897, 0.596729706740007, 0.476327806478366, 0.509105541277677, 0.132755615515634, 0.79560792259872, 0.844249506713822, 0.852449066238478, 0.614817695692182, 0.194548760307953, 0.276970969513059, 0.298617821652442, 0.0162338875234127, 0.95308379875496, 0.47322135954164, 0.354269130621105, 0.941423683427274, 0.852623045444489, 0.754432768560946, 0.931352832121775, 0.169991446658969, 0.106862558284774, 0.971970159327611, 0.144399283220991, 0.351859032176435, 0.935040743788704, 0.92673868057318, 0.586484697880223, 0.276902898913249, 0.746454028878361, 0.890186936361715, 0.11445689224638, 0.328725935425609, 0.714610485592857, 0.246908984612674, 0.648443329380825, 0.351052222773433, 0.0287432507611811, 0.734135948354378, 0.194968307157978, 0.955393396550789, 0.20379597810097, 0.767266945680603, 0.486582137644291, 0.487953958334401, 0.269033615011722, 0.206794039579108, 0.917672138195485, 0.999826987972483, 0.201813277089968, 0.293077527312562, 0.488147204276174, 0.0925062848255038, 0.323177747661248, 0.163101071026176, 0.814244165318087, 0.342218021629378, 0.131897852756083, 0.50597850070335, 0.693065346451476, 0.364949073409662, 0.90457378490828, 0.0911725244950503, 0.413594011915848, 0.335864054271951, 0.746740082046017, 0.8374565965496, 0.848530902992934, 0.937376967165619, 0.841480806469917, 0.312973105581477, 0.323705927468836, 0.227601936785504, 0.419034202350304, 0.711184817831963, 0.680461543146521, 0.505369338905439, 0.517901089275256, 0.513440987095237, 0.555394522612914, 0.938140297774225, 0.788064481923357, 0.292707462795079, 0.317566408542916, 0.0982909849844873, 0.0911575744394213, 0.758282830240205, 0.0953669759910554, 0.250597510486841, 0.572214664891362, 0.0935809188522398, 0.145962286973372, 0.855308949947357, 0.211951365228742, 0.998465299606323, 0.960390452994034, 0.756183248944581, 0.891742505133152, 0.137093368452042, 0.177841594209895, 0.105426595779136, 0.879545402945951, 0.406723195686936, 0.974694820120931, 0.630447847768664, 0.790671613067389, 0.626305991318077, 0.133715802803636, 0.435572562506422, 0.110036046709865, 0.958487988449633, 0.0429108676034957, 0.509250744711608, 0.468056805664673, 0.290612827986479, 0.78608355903998, 0.841051378753036, 0.279053965583444, 0.457730471622199, 0.00464116060175002, 0.0992977067362517, 0.556961689144373, 0.715104911942035, 0.0117636520881206, 0.946843601297587, 0.43208628334105, 0.727391508640721, 0.170340347336605, 0.535907811718062, 0.401854709489271, 0.443134960951284, 0.728855505120009, 0.226727265398949, 0.644256435567513, 0.519909818656743, 0.977409548126161, 0.668394329724833, 0.34937855321914, 0.1484789038077, 0.958692349959165, 0.761515629477799, 0.523122211918235, 0.25198934879154, 0.715475084027275, 0.11482755606994, 0.403458928689361, 0.504769472638145, 0.891740666003898, 0.559222171315923, 0.209492321126163, 0.45837097405456, 0.823475934332237, 0.138019165024161, 0.650525250704959, 0.464787728851661, 0.593120660167187, 0.970062642591074, 0.929372897837311, 0.908305455930531, 0.910642504226416, 0.431640760973096, 0.19285691366531, 0.595147019950673, 0.270145718706772, 0.876931878970936, 0.552513117203489, 0.617030482972041, 0.539893196430057, 0.579720962559804, 0.47551919403486, 0.911328823771328, 0.70484364614822, 0.158823720412329, 0.195010740542784, 0.00683005014434457, 0.591234432300553, 0.190671156859025, 0.587180166505277, 0.301195236621425, 0.323616542620584, 0.985875635175034, 0.0433837936725467, 0.364959573373199, 0.654061825945973, 0.475679791765288, 0.0012798672541976, 0.106855802470818, 0.257363406708464, 0.980263154720888, 0.396348708542064, 0.473302874946967, 0.397903566248715, 0.879134934395552, 0.706742078997195, 0.763938858639449, 0.432955668307841, 0.884701420553029, 0.858070468530059, 0.675592797575518, 0.81629575858824, 0.245757478522137, 0.344568273285404, 0.778321781661361, 0.392228118143976, 0.822391235968098, 0.029155929107219, 0.333642615005374, 0.758283053524792, 0.699138527736068, 0.192171358736232, 0.628946361830458, 0.0970558719709516, 0.0805424032732844, 0.116309506585822, 0.256178668700159, 0.16028324351646, 0.120499612996355, 0.153013636590913, 0.568948318716139, 0.295015802141279, 0.589719495037571, 0.541657338384539, 0.00939681846648455, 0.770411866949871, 0.231701928889379, 0.634875555289909, 0.657819025684148, 0.331425155280158, 0.299486440140754, 0.347160363337025, 0.245661673136055, 0.0894057138357311, 0.408513325033709, 0.623279398307204, 0.0536627969704568, 0.494568474357948, 0.662609312916175, 0.135042963316664, 0.603881729766726, 0.501098241424188, 0.63601370761171, 0.300563159631565, 0.774108817800879, 0.358125010505319, 0.0700125463772565, 0.0556461052037776, 0.485355453565717, 0.572960356716067, 0.119145602919161, 0.865947283571586, 0.820702457334846, 0.903463252121583, 0.0470460380893201, 0.708595262374729, 0.0567074478603899, 0.246073011308908, 0.509116917150095, 0.883285562274978, 0.408168029040098, 0.602504180977121, 0.246997465612367, 0.597427988890558, 0.863945377292112, 0.593564048875123, 0.178691865643486, 0.0234932953026146, 0.708725963253528, 0.0712701927404851, 0.536721953423694, 0.357776971068233, 0.491801067953929, 0.880017036106437, 0.640617923578247, 0.278831052361056, 0.307965766405687, 0.0563153568655252, 0.404552730731666, 0.630832699825987, 0.551601600134745, 0.888928052736446, 0.964540429180488, 0.00065249134786427, 0.72478033718653, 0.688024952309206, 0.73778010928072, 0.620786507381126, 0.311322793830186, 0.0122596640139818, 0.0634456772822887, 0.119673076551408, 0.857474458403885, 0.792895857710391, 0.90095031587407, 0.188773279311135, 0.105878230184317, 0.514575008768588, 0.049144443590194, 0.10439943196252, 0.3632660608273, 0.283068237127736, 0.360461977543309, 0.933702779468149, 0.361754542216659, 0.837993694934994, 0.898243188858032, 0.408389635616913, 0.422743043163791, 0.368805320002139, 0.98927751602605, 0.182272276375443, 0.0870709756854922, 0.677231072215363, 0.612478002673015, 0.817585434298962, 0.0332216802053154, 0.955149207264185, 0.688669318566099, 0.402499172370881, 0.968147807288915, 0.384715412044898, 0.851869379868731, 0.98499821103178, 0.292654942022637, 0.0568247376941144, 0.150192786939442, 0.00459100259467959, 0.911822159774601, 0.179351090686396, 0.789012154564261, 0.288535422412679, 0.802349874516949, 0.59324937290512, 0.0781319404486567, 0.370988622074947, 0.203630216186866, 0.804933947278187, 0.131571124540642, 0.828524231212214, 0.848276543430984, 0.953145212261006, 0.0853757329750806, 0.21092888340354, 0.672334382077679, 0.920054770540446, 0.35443621664308, 0.554193252231926, 0.0416994553525001, 0.472861256683245, 0.935308087151498, 0.491899880114943, 0.295799763640389, 0.798121861182153, 0.046311120968312, 0.854258546838537, 0.298425349639729, 0.342193471267819, 0.544778716051951, 0.168349597835913, 0.89474160131067, 0.779031608719379, 0.240121629554778, 0.405683236662298, 0.040540921734646, 0.975231247954071, 0.30638601956889, 0.111086843535304, 0.103246744256467, 0.176043132785708, 0.36730745784007, 0.995085651986301, 0.302161384606734, 0.31705296295695, 0.159387059044093, 0.106668738182634, 0.63698944542557, 0.470090374117717, 0.553280318621546, 0.410580684430897, 0.722563271177933, 0.219424898736179, 0.949964464874938, 0.956017588498071, 0.0734421645756811, 0.245965385809541, 0.0598306853789836, 0.315521084470674, 0.464948303066194, 0.732373290928081, 0.885765784885734, 0.925915576051921, 0.868177025811747, 0.401186160044745, 0.820773157291114, 0.802928718505427, 0.378386207856238, 0.333107637474313, 0.658316427608952, 0.231045863358304, 0.428868243470788, 0.847250893246382, 0.590510870562866, 0.743681150721386, 0.739821596769616, 0.198897932190448, 0.0229621774051338, 0.40441044094041, 0.4538688310422, 0.0370801337994635, 0.586796198273078, 0.85114680766128, 0.466510647675022, 0.72125727776438, 0.597850512480363, 0.970445530256256, 0.824383437866345, 0.915100604528561, 0.78894290747121, 0.0760166840627789, 0.446872147498652, 0.113246088381857, 0.291081183357164, 0.686134767951444, 0.0177396505605429, 0.204887747066095, 0.960173872997984, 0.641550589585677, 0.126762071857229, 0.982202292187139, 0.503218550235033, 0.620918002445251, 0.534736085683107, 0.0415312019176781, 0.317106332164258, 0.708187042502686, 0.637623490765691, 0.548848294187337, 0.623919502366334, 0.456282211234793, 0.204802228603512, 0.970296283252537, 0.19097574823536, 0.480038233101368, 0.434966658474877, 0.787358567351475, 0.41331087751314, 0.457242771983147, 0.489831270184368, 0.113760035019368, 0.721844929503277, 0.0929107423871756, 0.355879611335695, 0.573024924378842, 0.233123685931787, 0.6429634608794, 0.58679801505059, 0.108658795244992, 0.795727781252936, 0.0129886064678431, 0.350833832519129, 0.760371719952673, 0.317734786542132, 0.114384569693357, 0.337700912263244, 0.0274328498635441, 0.205833068815991, 0.31276917620562, 0.921513265231624, 0.889448454603553, 0.721953355707228, 0.208532121032476, 0.153553274227306, 0.0955133135430515, 0.969104799209163, 0.371698021423072, 0.47394798672758, 0.807637023041025, 0.385256400099024, 0.986533994786441, 0.73309546103701, 0.987721901619807, 0.145026509650052, 0.643101030495018, 0.854426043573767, 0.73421481042169, 0.566274993587285, 0.447484823409468, 0.479175786254928, 0.958860090468079, 0.376546438317746, 0.312345372978598, 0.572406751802191, 0.238745126174763, 0.117383589502424, 0.803579399595037, 0.99833844252862, 0.0596549401525408, 0.483671794412658, 0.438169909873977, 0.291023748693988, 0.13914058660157, 0.0352492788806558, 0.605193154420704, 0.30764270760119, 0.443740031681955, 0.231638493714854, 0.21970993373543, 0.417796609224752, 0.32400220585987, 0.0657286588102579, 0.401510232128203, 0.153471017256379, 0.125627531204373, 0.540871791308746, 0.665477641392499, 0.0461612367071211, 0.200053315842524, 0.400631225900725, 0.929336148081347, 0.910069790668786, 0.125654823379591, 0.538420549128205, 0.522940589580685, 0.0801268855575472, 0.984736381564289, 0.25519923097454, 0.0823310809209943, 0.00352790975011885, 0.687174743739888, 0.0306961245369166, 0.664772119838744, 0.525875269668177, 0.207540407078341, 0.204068706138059, 0.456610308028758, 0.144474439090118, 0.259537893114612, 0.652829773025587, 0.968382609542459, 0.576939217746258, 0.991258810739964, 0.562992702936754, 0.821124869864434, 0.856978189200163, 0.219969375990331, 0.0967946688178927, 0.569075882667676, 0.938001617789268, 0.0727311582304537, 0.758298781234771, 0.788740180898458, 0.942300020717084, 0.064331135712564, 0.186737353680655, 0.460437160218135, 0.38172340230085, 0.0518650289159268, 0.348242917563766, 0.540382326114923, 0.99906910257414, 0.883083375869319, 0.55731227947399, 0.367713082581758, 0.733122798381373, 0.463479903293774, 0.968267790507525, 0.0540189852472395, 0.219449118245393, 0.596686239121482, 0.914992895908654, 0.0596107235178351, 0.19398624333553, 0.756860638037324, 0.0756060013081878, 0.13581521785818, 0.653700879309326, 0.102268222952262, 0.413463892647997, 0.529032219201326, 0.677323610754684, 0.122678809100762, 0.902922067092732, 0.791056103771552, 0.0187505162321031, 0.279111265670508, 0.726207613945007, 0.224383761640638, 0.142573039745912, 0.970985303167254, 0.0424321482423693, 0.012608585646376, 0.25158839370124, 0.514548631152138, 0.769435291644186, 0.0242532584816217, 0.397158089792356, 0.056611803593114, 0.671101413201541, 0.736261212034151, 0.0939119651447982, 0.0483818624634296, 0.229521757923067, 0.560090633342043, 0.04589913925156, 0.206212881952524, 0.11469554505311, 0.854652147972956, 0.586148975649849, 0.964731245767325, 0.179904066259041, 0.714506753254682, 0.359421872999519, 0.682132095098495, 0.312679512426257, 0.819836052367464, 0.991861471673474, 0.691151596838608, 0.0692041215952486, 0.315528010251001, 0.0416791974566877, 0.937565565574914, 0.809507315047085, 0.0924601883161813, 0.749493284150958, 0.895238177850842, 0.36146730161272, 0.270662652561441, 0.832635543309152, 0.251541799167171, 0.313660310115665, 0.35650225635618, 0.236423360416666, 0.618874595034868, 0.0936965618748218, 0.308583118254319, 0.627899456070736, 0.226935843937099, 0.906728325178847, 0.887167657958344, 0.412744249217212, 0.863601594697684, 0.276832292322069, 0.478959905682132, 0.0330305139068514, 0.88363968138583, 0.00066221971064806, 0.491635447600856, 0.829682876588777, 0.810609807725996, 0.221550068119541, 0.238247471628711, 0.388227749150246, 0.976448472589254, 0.271393097704276, 0.321394196012989, 0.802997623104602, 0.916381586110219, 0.666128802113235, 0.109419248998165, 0.272137602325529, 0.468897260725498, 0.374398417770863, 0.761538264574483, 0.0191268315538764, 0.917395962402225, 0.295621898258105, 0.970473116729409, 0.0986033827066422, 0.866336778970435, 0.324422063538805, 0.678594386205077, 0.2806897519622, 0.337597959442064, 0.34208916942589, 0.0865799356251955, 0.00843900418840349, 0.18699528533034, 0.366880461107939, 0.452148049604148, 0.537430062424392, 0.120529194362462, 0.295923253521323, 0.576014467282221, 0.514129482442513, 0.129231121158227, 0.692385851638392, 0.849711124552414, 0.246567967580631, 0.46199849108234, 0.686702027218416, 0.418075525434688, 0.740315810078755, 0.026057634036988, 0.108733740169555, 0.672478249995038, 0.147432605503127, 0.520205376436934, 0.055008151801303, 0.633748637977988, 0.0436463064979762, 0.375926251756027, 0.0545277989003807, 0.114993437426165, 0.852713093394414, 0.988664999604225, 0.383526071440428, 0.713219781871885, 0.0449140323325992, 0.972840188071132, 0.449917468708009, 0.930675677256659, 0.798653110396117, 0.593218411551788, 0.859244636958465, 0.910249582258984, 0.330451243091375, 0.615313008660451, 0.0743392307776958, 0.739047133130953, 0.726365269161761, 0.74413991975598, 0.36693831672892, 0.185802943073213, 0.662006915779784, 0.640157556626946, 0.721927968785167, 0.624938840745017, 0.677186945686117, 0.998518635053188, 0.901145257987082, 0.846272112801671, 0.116481346078217, 0.118878022069111, 0.410612330539152, 0.490591303445399, 0.747500386321917, 0.474561055889353, 0.463490088703111, 0.755631801206619, 0.404877866385505, 0.344914654968306, 0.560520028928295, 0.663199326954782, 0.457307873060927, 0.0918332473374903, 0.776322023011744, 0.965696285478771, 0.551742364186794, 0.149342010496184, 0.069498797878623, 0.305513724684715, 0.556854054331779, 0.786492242943496, 0.0562246874906123, 0.0935100971255451, 0.0146542687434703, 0.420901159290224, 0.783991944277659, 0.906043791677803, 0.305222465656698, 0.819286513142288, 0.478123124688864, 0.643579783616588, 0.954174470854923, 0.531108044786379, 0.380217537749559, 0.317691540112719, 0.714768592035398, 0.595648543443531, 0.999444995075464, 0.660466268425807, 0.689843739150092, 0.986922975862399, 0.460092191351578, 0.473807683913037, 0.929681024746969, 0.261779008898884, 0.63973848009482, 0.769424800993875, 0.483653571922332, 0.759387714089826, 0.430044373963028, 0.452683693729341, 0.325632978463545, 0.587344500236213, 0.30936914589256, 0.0773299445863813, 0.629365648375824, 0.341049629962072, 0.295499305007979, 0.481661309720948, 0.843421611003578, 0.75511080538854, 0.698104239301756, 0.152487272396684, 0.286448122235015, 0.0318267813418061, 0.0914786257781088, 0.624168718466535, 0.120850498788059, 0.673865277087316, 0.331238167593256, 0.829900992801413, 0.120651193195954, 0.191918866243213, 0.727713590720668, 0.568511850899085, 0.509355500806123, 0.378573850030079, 0.340504657709971, 0.513743012910709, 0.0877056701574475, 0.156008960213512, 0.810526253888384, 0.877127019222826, 0.0418912526220083, 0.459223999641836, 0.28793776081875, 0.311749001732096, 0.028414860367775, 0.0571856750175357, 0.558781824773178, 0.440266399411485, 0.152943632798269, 0.523176665650681, 0.833073200192302, 0.30819331551902, 0.211539538344368, 0.0866923804860562, 0.650751485023648, 0.844217300415039, 0.198224845575169, 0.777236570836976, 0.299947881372645, 0.429145447677001, 0.189137258101255, 0.709031788399443, 0.937194138998166, 0.0362644421402365, 0.0577078156638891, 0.0179103484842926, 0.978917961940169, 0.81410500803031, 0.251690958393738, 0.940680927829817, 0.539572800509632, 0.67807694291696, 0.540880476590246, 0.885833704378456, 0.859725511400029, 0.344662874005735, 0.266714124940336, 0.984361508628353, 0.547212481265888, 0.521131195127964, 0.109392086276785, 0.757000288227573, 0.953091950854287, 0.294771308312193, 0.419969449751079, 0.850184505339712, 0.156091944081709, 0.839800185058266, 0.335157239343971, 0.55564552731812, 0.0990244532003999, 0.304167989874259, 0.00604521855711937, 0.692521496675909, 0.448132474208251, 0.727701645810157, 0.74535027006641), actual = c(0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"