Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>       0.00000000       1.00000000       0.98912529       0.08495014 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>       0.50000000       0.49896653       0.49896653       0.50583248 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.752639399375767, 0.670487194089219, 0.281300383852795, 0.130222169915214, 0.341387113789096, 0.600176272448152, 0.949894406134263, 0.041673289379105, 0.803147141588852, 0.561979870544747, 0.141166935209185, 0.228321446571499, 0.937669339124113, 0.385814704466611, 0.0323692911770195, 0.164809166686609, 0.305891156196594, 0.022083151852712, 0.770428233314306, 0.628538907505572, 0.953149585286155, 0.406197325326502, 0.338575160130858, 0.298083994537592, 0.934210556792095, 0.0267232458572835, 0.665850190911442, 0.67007432016544, 0.00296675460413098, 0.131763440556824, 0.671902359696105, 0.937658796785399, 0.303256968501955, 0.495552681852132, 0.906690306030214, 0.582379270344973, 0.581437822664157, 0.102287834510207, 0.313766664592549, 0.242733595892787, 0.722917499486357, 0.214898945065215, 0.328002809081227, 0.517506760777906, 0.49194603972137, 0.432402255246416, 0.159579979721457, 0.86868727975525, 0.342309476574883, 0.992435775464401, 0.996506893541664, 0.45079512312077, 0.488442551577464, 0.135601873509586, 0.966906116809696, 0.379135868744925, 0.407939344178885, 0.341132346075028, 0.814633947098628, 0.999198351288214, 0.858941996470094, 0.685070924926549, 0.100877019809559, 0.43447710829787, 0.341787742683664, 0.656493169721216, 0.651217815000564, 0.0120811003725976, 0.943135207053274, 0.91171542624943, 0.349456339608878, 0.562897408846766, 0.940553560620174, 0.507814499083906, 0.744770068209618, 0.0228968297597021, 0.470206181751564, 0.341543664690107, 0.279093347955495, 0.168295268667862, 0.254426365718246, 0.826549777528271, 0.450723505113274, 0.92573261866346, 0.449087293585762, 0.319823771947995, 0.0962896810378879, 0.256964009720832, 0.914619564078748, 0.352528206538409, 0.367407848360017, 0.152984760701656, 0.388286961941049, 0.537661156617105, 0.917407994624227, 0.995644975919276, 0.960606078617275, 0.539306605933234, 0.667558541288599, 0.577105550095439, 0.795069773914292, 0.943654597038403, 0.340842685895041, 0.786339059472084, 0.431241115322337, 0.209499041317031, 0.962140214862302, 0.652652223361656, 0.224295645952225, 0.579530534334481, 0.725619956851006, 0.700161260319874, 0.79124231194146, 0.249143702210858, 0.647274284623563, 0.469508352689445, 0.625563041539863, 0.739982316736132, 0.524394062114879, 0.700866517145187, 0.288513333303854, 0.332611123332754, 0.284754341933876, 0.422711673658341, 0.372780847828835, 0.474735418800265, 0.255795890232548, 0.791655078297481, 0.292718080803752, 0.0849496282171458, 0.353909112978727, 0.522281021578237, 0.666457967599854, 0.64201651699841, 0.475138475187123, 0.037935156840831, 0.428296024678275, 0.61301540536806, 0.651716294931248, 0.952402913477272, 0.00735076237469912, 0.962326053529978, 0.120955768972635, 0.937061817152426, 0.0809083918575197, 0.0690264538861811, 0.316273639909923, 0.234469115268439, 0.472390302456915, 0.955554488347843, 0.561038328334689, 0.422353708185256, 0.622376667801291, 0.336620845133439, 0.982890357263386, 0.970182640478015, 0.58332241512835, 0.294682912062854, 0.313552015461028, 0.406186298467219, 0.963123076828197, 0.681272827554494, 0.791986839147285, 0.138815019512549, 0.205270829843357, 0.137383246794343, 0.14099826477468, 0.365144328679889, 0.988593842368573, 0.712213365128264, 0.289869762258604, 0.840858286013827, 0.472636686405167, 0.905986992642283, 0.855684850597754, 0.419713917421177, 0.660511641995981, 0.9801047204528, 0.720672300551087, 0.314098968869075, 0.0440349066630006, 0.408078217878938, 0.51848640688695, 0.460513570345938, 0.440120159648359, 0.431134617188945, 0.00837492221035063, 0.506682748207822, 0.942105717025697, 0.771471517160535, 0.139709623763338, 0.864870437420905, 0.957748077809811, 0.886073588160798, 0.685892647365108, 0.164478683145717, 0.816539677092806, 0.526543970219791, 0.98504661116749, 0.0800920475739986, 0.0549228494055569, 0.402059714775532, 0.58948381152004, 0.493165051797405, 0.553192828549072, 0.628269309643656, 0.273733081528917, 0.407929466804489, 0.0567747850436717, 0.72128120623529, 0.619999985676259, 0.60654287179932, 0.801461264491081, 0.284944960847497, 0.414452739758417, 0.553203224902973, 0.929496959783137, 0.961411150870845, 0.716358424397185, 0.0396553806494921, 0.0161784838419408, 0.0762556591071188, 0.118612074526027, 0.76953536667861, 0.13379472005181, 0.822618682635948, 0.61553322756663, 0.0067519168369472, 0.363409559242427, 0.469745892798528, 0.644935687072575, 0.339442745083943, 0.571684398455545, 0.328111388953403, 0.920606434345245, 0.68827533419244, 0.973439159803092, 0.791066954610869, 0.131254089064896, 0.866422896040604, 0.52276096586138, 0.0176935149356723, 0.848123799310997, 0.866553041618317, 0.542456336552277, 0.500259292311966, 0.608907906571403, 0.125435691559687, 0.127756389789283, 0.394176388392225, 0.42603746149689, 0.726238908013329, 0.825939716538414, 0.215932104503736, 0.932720075827092, 0.536351287737489, 0.541699219262227, 0.661008231109008, 0.61916091828607, 0.209875106811523, 0.782483709976077, 0.147244418272749, 0.705201026517898, 0.710235578939319, 0.0254908283241093, 0.932213662890717, 0.783523286925629, 0.898307644529268, 0.927203090628609, 0.318494476610795, 0.978174021001905, 0.286637076875195, 0.352035386255011, 0.584919230313972, 0.975565206026658, 0.461087197763845, 0.494190222816542, 0.921524869278073, 0.955639380728826, 0.204333610832691, 0.991093745222315, 0.158104752423242, 0.417143275728449, 0.455890106037259, 0.612547531025484, 0.0132512249983847, 0.379673899849877, 0.0421331331599504, 0.442071734694764, 0.450804633554071, 0.00354922725819051, 0.440264710457996, 0.775488371495157, 0.673474353505298, 0.264015307882801, 0.982979891356081, 0.643577350303531, 0.355909844394773, 0.159017361933365, 0.241248219041154, 0.705470028333366, 0.214678653283045, 0.651389192324132, 0.0414673844352365, 0.437435865867883, 0.734233155846596, 0.668559826212004, 0.82973309000954, 0.654773851390928, 0.369381338125095, 0.727964316261932, 0.915650688810274, 0.888657800620422, 0.308674142695963, 0.693752486491576, 0.731720601208508, 0.233327475842088, 0.240781465079635, 0.706054551526904, 0.0819993105251342, 0.000293491873890162, 0.96805024635978, 0.414847085718066, 0.696438315324485, 0.324197727954015, 0.710501812631264, 0.211486684158444, 0.858565881382674, 0.82511711330153, 0.148709001019597, 0.521708362037316, 0.59128790628165, 0.176899350481108, 0.20547794830054, 0.737083756830543, 0.670680948533118, 0.129749221261591, 0.610040328232571, 0.12266349606216, 0.598253816599026, 0.275607385206968, 0.514057430438697, 0.332086188253015, 0.907863059779629, 0.67176428087987, 0.604152518324554, 0.856448377482593, 0.795404731761664, 0.947290813317522, 0.990257303463295, 0.547943113138899, 0.296079874038696, 0.882673697080463, 0.130733674624935, 0.950675946194679, 0.428301183041185, 0.727595276664943, 0.51851913984865, 0.12285264278762, 0.968963449355215, 0.733792376006022, 0.674777041189373, 0.168612678069621, 0.946671447483823, 0.907890753587708, 0.123723914381117, 0.433686728356406, 0.727892671246082, 0.855370338540524, 0.805701243691146, 0.327996539417654, 0.492341217584908, 0.321485325461254, 0.112652227049693, 0.648275145329535, 0.0368314029183239, 0.121931673027575, 0.865313764661551, 0.988852915121242, 0.563385582296178, 0.516368820564821, 0.15141798555851, 0.965153437107801, 0.83681576885283, 0.884015412768349, 0.14687750954181, 0.69220191729255, 0.225022080587223, 0.572208505822346, 0.387791735352948, 0.681146025191993, 0.3482875963673, 0.748458773829043, 0.146395747084171, 0.531310925260186, 0.608956013806164, 0.0112803210504353, 0.570963520556688, 0.850770174758509, 0.394031598931178, 0.625193175859749, 0.891660680994391, 0.702511071925983, 0.0862240088172257, 0.500579097541049, 0.48017045063898, 0.66815850418061, 0.304079577326775, 0.772916377987713, 0.581265227636322, 0.207804243080318, 0.635707878740504, 0.18768021534197, 0.866738194599748, 0.319520259741694, 0.457178494427353, 0.662832492496818, 0.593113843817264, 0.853564376011491, 0.714750797953457, 0.43936935858801, 0.18363696965389, 0.409715969348326, 0.356352012604475, 0.0112407710403204, 0.86624318594113, 0.415299066575244, 0.403500362066552, 0.830705671105534, 0.606479330919683, 0.628009942127392, 0.288800252601504, 0.896108428947628, 0.323331253137439, 0.894437913084403, 0.359788167988881, 0.604912696406245, 0.523023534100503, 0.912382860668004, 0.877631508745253, 0.18462487612851, 0.258738664211705, 0.73799906973727, 0.321670146659017, 0.340827445499599, 0.72604336310178, 0.392199534457177, 0.170815044548362, 0.834878842579201, 0.799198722001165, 0.398064533248544, 0.217980665853247, 0.0202881342265755, 0.576983879087493, 0.87221991806291, 0.584011441329494, 0.477084800833836, 0.140660492237657, 0.0461861870717257, 0.21947006159462, 0.227678932948038, 0.798198003554717, 0.295014796080068, 0.495274230837822, 0.883762953570113, 0.0126084559597075, 0.396376764168963, 0.502221870003268, 0.25550307193771, 0.883923730347306, 0.939419906819239, 0.633938619634137, 0.385209230240434, 0.807086967397481, 0.944362414302304, 0.579209140269086, 0.319955843500793, 0.719626255799085, 0.522621189709753, 0.841157793765888, 0.589013710385188, 0.825365060707554, 0.865383679512888, 0.938376462087035, 0.646014313213527, 0.627313852543011, 0.0666671514045447, 0.480873321415856, 0.0187761997804046, 0.077825850341469, 0.0322722150012851, 0.594646808924153, 0.627890351694077, 0.855070613091812, 0.940653232391924, 0.442015429027379, 0.0171263839583844, 0.519065745174885, 0.512313174782321, 0.103127542417496, 0.673198527889326, 0.43873578030616, 0.976149634225294, 0.234276394126937, 0.264140287414193, 0.878774218959734, 0.714812563732266, 0.576429185923189, 0.839873854070902, 0.00482252216897905, 0.782276920741424, 0.423334313090891, 0.845444986131042, 0.387273349100724, 0.0454992854502052, 0.168263859581202, 0.763270172057673, 0.0322136275935918, 0.441851432900876, 0.0121853637974709, 0.63807710702531, 0.256683832267299, 0.620055651525036, 0.538826060248539, 0.169754539383575, 0.576170652871951, 0.492151155602187, 0.0150406316388398, 0.471629659645259, 0.857474972028285, 0.00351054407656193, 0.683632754487917, 0.430803679628298, 0.821025139885023, 0.6789135474246, 0.0880562614183873, 0.383807509671897, 0.0403030631132424, 0.682863899040967, 0.781702946638688, 0.907863784348592, 0.857420555315912, 0.126710176002234, 0.937826552661136, 0.829768779221922, 0.0345261667389423, 0.775747699895874, 0.420702534494922, 0.545877408003435, 0.88519606157206, 0.541356496978551, 0.662231175927445, 0.0792593131773174, 0.165008722338825, 0.958518485073, 0.281090252334252, 0.99236951675266, 0.188271491788328, 0.508545630611479, 0.63561708573252, 0.903193149250001, 0.993304750416428, 0.00671388464979827, 0.952849288238212, 0.597978392383084, 0.726877378998324, 0.749462860403582, 0.545975085813552, 0.3202347853221, 0.967814025469124, 0.70076806563884, 0.201106971595436, 0.710509026888758, 0.00660832785069942, 0.493256552610546, 0.860176287358627, 0.363155728206038, 0.474284904776141, 0.153709637466818, 0.319032145896927, 0.280543386004865, 0.470078167505562, 0.483006500406191, 0.17966717062518, 0.685881829122081, 0.00359584391117096, 0.64330839109607, 0.436892824014649, 0.128706111572683, 0.567898939130828, 0.960331041598693, 0.385519278002903, 0.360715919407085, 0.646013310644776, 0.916300120996311, 0.585386104881763, 0.187396670226008, 0.555944405728951, 0.532216941704974, 0.87047604820691, 0.0349178346805274, 0.960479151224717, 0.506480482639745, 0.20399402314797, 0.193268615752459, 0.391489742323756, 0.388469154480845, 0.570535158971325, 0.281329035526142, 0.826059699757025, 0.242812672164291, 0.793538682861254, 0.595610432559624, 0.439573612064123, 0.460258034523576, 0.042644084431231, 0.156371923396364, 0.721179314190522, 0.353227979270741, 0.67770867375657, 0.647139472188428, 0.7401611241512, 0.976637011859566, 0.442164445528761, 0.133116058539599, 0.028548939852044, 0.196820396697149, 0.0177951799705625, 0.230292199645191, 0.779364501824602, 0.726014723768458, 0.388540806947276, 0.70490558934398, 0.126588228391483, 0.490684288321063, 0.837159119779244, 0.622410960728303, 0.875567490933463, 0.363745017908514, 0.837633094284683, 0.280305665684864, 0.366607828298584, 0.507310193264857, 0.0313140286598355, 0.480280343210325, 0.0145730969961733, 0.352360201068223, 0.904782136902213, 0.586479281308129, 0.531849897000939, 0.801979398354888, 0.796631439588964, 0.017518118256703, 0.370712459785864, 0.946552209556103, 0.8887030670885, 0.959822409320623, 0.0710553997196257, 0.841508052079007, 0.915661918930709, 0.334159960271791, 0.480608920101076, 0.608259028755128, 0.129187177168205, 0.378608654718846, 0.570780917303637, 0.890599365346134, 0.748211837606505, 0.365066779078916, 0.551725880941376, 0.76500984816812, 0.469912890577689, 0.735168921295553, 0.524669085396454, 0.909735459368676, 0.502955528441817, 0.621041932608932, 0.521034274948761, 0.505969444755465, 0.356112946756184, 0.331862695282325, 0.391807822044939, 0.489165607606992, 0.761145680677146, 0.87636621017009, 0.570327647496015, 0.360218192450702, 0.00274737877771258, 0.483584772096947, 0.332064314512536, 0.283674547914416, 0.0127152064815164, 0.240073659457266, 0.223631806205958, 0.588012855965644, 0.898891426855698, 0.00794478110037744, 0.304843179415911, 0.273871736368164, 0.774760222295299, 0.162778968224302, 0.883546612923965, 0.0142008534166962, 0.234339587623253, 0.335529235424474, 0.476752880727872, 0.71723630046472, 0.126648190198466, 0.593567876610905, 0.686218491988257, 0.00374262291006744, 0.962097502546385, 0.215062740957364, 0.311697523575276, 0.111297076800838, 0.0904542275238782, 0.965350446058437, 0.609658770496026, 0.568875053431839, 0.450081187533215, 0.438969852868468, 0.669596188003197, 0.670354213798419, 0.327819642145187, 0.953956406097859, 0.293466777773574, 0.897404741728678, 0.221566768363118, 0.846807357622311, 0.267930330010131, 0.718818880384788, 0.950481429463252, 0.784104749793187, 0.924450160237029, 0.405011984985322, 0.365323004778475, 0.118727537570521, 0.10832404345274, 0.26542538870126, 0.436914184130728, 0.453610127558932, 0.307375350268558, 0.810689651640132, 0.416972370585427, 0.881038377294317, 0.841833431506529, 0.148069025017321, 0.739671958843246, 0.942752496805042, 0.324432291323319, 0.0669647615868598, 0.731768217636272, 0.411029784008861, 0.121429282939062, 0.640220484696329, 0.970418818527833, 0.491428369423375, 0.864220099756494, 0.156349926488474, 0.27070544147864, 0.895617469213903, 0.0693352934904397, 0.55943017010577, 0.103801540099084, 0.883184263948351, 0.779992349911481, 0.844244859414175, 0.170256223063916, 0.873559015337378, 0.070587390800938, 0.120896466774866, 0.529237287351862, 0.321554675931111, 0.405832574702799, 0.730031171580777, 0.0947496963199228, 0.267210924299434, 0.0845849195029587, 0.636632841778919, 0.0752050231676549, 0.958156760549173, 0.339014393044636, 0.594996250234544, 0.541097774170339, 0.745595423970371, 0.512209660373628, 0.000811456935480237, 0.0410935422405601, 0.521897864993662, 0.607247925829142, 0.788557280553505, 0.184702704660594, 0.356972660403699, 0.906050013378263, 0.224962041946128, 0.189514979720116, 0.597287102602422, 0.791795321041718, 0.261245764093474, 0.492010441375896, 0.332045610062778, 0.757660760777071, 0.288028070237488, 0.637852329062298, 0.732078669592738, 0.243758597178385, 0.65797670930624, 0.415334431221709, 0.423998401500285, 0.238672327483073, 0.955974710173905, 0.388248491566628, 0.930003165500239, 0.422911692876369, 0.0262250551022589, 0.578522747848183, 0.349250260274857, 0.516021018382162, 0.00319047714583576, 0.578343517845497, 0.310394330183044, 0.232304310658947, 0.406666517956182, 0.572094129631296, 0.413031285861507, 0.666862770682201, 0.169830994913355, 0.361455703619868, 0.822229319950566, 0.345915880752727, 0.949816177133471, 0.286514886422083, 0.026909596985206, 0.453957094578072, 0.0952254405710846, 0.0814692855346948, 0.957323393318802, 0.502694121329114, 0.369080899516121, 0.438202315941453, 0.152483587386087, 0.862757897702977, 0.674258315237239, 0.124668108765036, 0.258175367023796, 0.342561802826822, 0.880456501850858, 0.749130238546059, 0.33328498667106, 0.0841667929198593, 0.589560497319326, 0.417918239254504, 0.303690232569352, 0.0921382275409997, 0.954917557304725, 0.290656078141183, 0.326989693101496, 0.673251599771902, 0.104421334108338, 0.478354123653844, 0.0752597176469862, 0.699915904086083, 0.176167803350836, 0.659156559500843, 0.771274257684126, 0.997819755692035, 0.888582267332822, 0.300255445297807, 0.439388245344162, 0.0253907849546522, 0.569472214207053, 0.305605017812923, 0.522385185118765, 0.969415789935738, 0.0969236758537591, 0.561610087985173, 0.868591288570315, 0.400382820982486, 0.15459298924543, 0.435771034564823, 0.385537586873397, 0.849195004906505, 0.715265514561906, 0.108118858654052, 0.681130877230316, 0.154462770326063, 0.419939626008272, 0.01966651994735, 0.301634642295539, 0.220109226647764, 0.43620349583216, 0.990091718733311, 0.543321018805727, 0.450994529062882, 0.234291625907645, 0.688779096351936, 0.99882584461011, 0.801592150237411, 0.429675069637597, 0.106767283054069, 0.193353237584233, 0.905506891896948, 0.721433282131329, 0.830099488608539, 0.783230164786801, 0.552233198424801, 0.428774093277752, 0.98987540579401, 0.621679550968111, 0.550770242232829, 0.638519678963348, 0.305792728438973, 0.184968084795401, 0.456725439988077, 0.0448501631617546, 0.487685755826533, 0.223205813672394, 0.796764399157837, 0.81988973217085, 0.197544787079096, 0.875861592125148, 0.228217513300478, 0.2993373577483, 0.113836590433493, 0.496304262429476, 0.732606330187991, 0.514862032374367, 0.139945360133424, 0.219916037749499, 0.373365636449307, 0.876036668429151, 0.820483408635482, 0.00526693393476307, 0.148684588959441, 0.201023783767596, 0.846829784335569, 0.756592262769118, 0.38816478708759, 0.773359185317531, 0.452763658715412, 0.266786836320534, 0.273300942033529, 0.937996495980769, 0.772946846438572, 0.555935427779332, 0.963971519144252, 0.950461542233825, 0.801125997211784, 0.169961533974856, 0.649379624286667, 0.742462404770777, 0.568147561512887, 0.889505188679323, 0.709612128557637, 0.273005158407614, 0.953607671195641, 0.0396753137465566, 0.111964842304587, 0.136926383944228, 0.792061918182299, 0.519706041784957, 0.836431560106575, 0.174501737579703, 0.156925652176142, 0.0593214295804501, 0.940128101035953, 0.101109337992966, 0.36800189758651, 0.471673925640061, 0.26202533300966, 0.200479505816475, 0.858220427529886, 0.382172773359343, 0.754179090959951, 0.0331238824874163, 0.0969671122729778, 0.18379718135111, 0.503942891489714, 0.76892926520668, 0.585107863647863, 0.69364608777687, 0.762604751856998, 0.561783927725628, 0.58376574027352, 0.170667151687667, 0.262297539273277, 0.89011811488308, 0.864660209976137, 0.1696583237499, 0.7573256874457, 0.821780093712732, 0.360515331383795, 0.656672768527642, 0.375590177020058, 0.772466453025118, 0.967727145878598, 0.924701454117894, 0.671965625602752, 0.00375090283341706, 0.0108575543854386), actual = c(0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.32079824921675, 0.97227617376484, 0.420017359778285, 0.0823255435097963, 0.122750197537243, 0.605800071032718, 0.970998853677884, 0.021230157231912, 0.235015469137579, 0.496605670079589, 0.567330167861655, 0.94980094069615, 0.632297381293029, 0.203678287798539, 0.826435155235231, 0.64248282299377, 0.433258703444153, 0.800100399181247, 0.760357209248468, 0.303572727367282, 0.13792651402764, 0.216635985998437, 0.158491609385237, 0.48463715147227, 0.171305261785164, 0.927893205080181, 0.686166256433353, 0.724469494773075, 0.900701356120408, 0.970437946263701, 0.433579000411555, 0.641636273823678, 0.696293526561931, 0.40432625846006, 0.554059690563008, 0.364341503707692, 0.412023642798886, 0.512043103808537, 0.695766163757071, 0.619715803069994, 0.807425167178735, 0.51423561386764, 0.0775092181283981, 0.746077405521646, 0.0462825610302389, 0.32722283154726, 0.173675331287086, 0.967477984027937, 0.646894951118156, 0.385714254109189, 0.328111049486324, 0.371495014987886, 0.773681095335633, 0.564174449304119, 0.0319937414024025, 0.456229409901425, 0.0819890112616122, 0.275673630880192, 0.45052547310479, 0.980856532929465, 0.0168154225684702, 0.30929722613655, 0.576374815311283, 0.177294870838523, 0.951317873550579, 0.909328780137002, 0.180023899301887, 0.978232473833486, 0.465680174995214, 0.901766892056912, 0.276695332257077, 0.41498187975958, 0.85444603045471, 0.404003501171246, 0.194709161296487, 0.916124913841486, 0.767251768615097, 0.917963183484972, 0.00610436801798642, 0.26730803353712, 0.659543660003692, 0.652898470172659, 0.435069612460211, 0.021730468608439, 0.802886858116835, 0.125373702729121, 0.930230964673683, 0.868416473735124, 0.318358079064637, 0.636113686021417, 0.456851076800376, 0.992889368673787, 0.70645986776799, 0.00408641481772065, 0.467777083162218, 0.457092033466324, 0.141147572314367, 0.161084789782763, 0.527158637763932, 0.652944800676778, 0.120020190021023, 0.300718969199806, 0.010383696295321, 0.754931300645694, 0.591306457528844, 0.692636073799804, 0.0492684321943671, 0.220970359398052, 0.333132558502257, 0.651667001191527, 0.484340792056173, 0.641602140385658, 0.635181539691985, 0.465943425893784, 0.535025604069233, 0.92656813072972, 0.305957864969969, 0.840886017307639, 0.338721290929243, 0.322746821679175, 0.684151046443731, 0.39315615571104, 0.573767110006884, 0.777162723708898, 0.506799592170864, 0.0477850290481001, 0.887023881776258, 0.626735333353281, 0.0102648709435016, 0.0847319907043129, 0.340814660303295, 0.118493292247877, 0.505725576542318, 0.659109118161723, 0.639351203339174, 0.320255645783618, 0.195098049705848, 0.995537516893819, 0.249004201497883, 0.771945603424683, 0.190174845978618, 0.293573826551437, 0.71028146147728, 0.536828515352681, 0.355941192246974, 0.130107211181894, 0.477792754769325, 0.549042322672904, 0.137495576869696, 0.97692073485814, 0.280061318771914, 0.211425329791382, 0.328781968913972, 0.532542815431952, 0.685124285053462, 0.132569467648864, 0.989158729091287, 0.334618260618299, 0.765316791133955, 0.660479534650221, 0.979669732740149, 0.950549689121544, 0.39519744575955, 0.500699530588463, 0.621784798568115, 0.452721311477944, 0.0728527843020856, 0.473524273373187, 0.56977225956507, 0.44378192233853, 0.0669614763464779, 0.00204006978310645, 0.587410148000345, 0.451798922149464, 0.794962037587538, 0.408685194794089, 0.349442089209333, 0.0151154429186136, 0.601719768252224, 0.587353364564478, 0.0710083525627851, 0.0124323009513319, 0.790521362097934, 0.0251446401234716, 0.386389285558835, 0.350079669617116, 0.598631336819381, 0.46996145346202, 0.730470984010026, 0.691383101744577, 0.992115743923932, 0.305196299916133, 0.144115916686133, 0.780036233831197, 0.944563534343615, 0.464717811206356, 0.220732759684324, 0.861491813091561, 0.386354709742591, 0.267788986675441, 0.743053070968017, 0.733333726646379, 0.701074359705672, 0.752251361962408, 0.994794271420687, 0.641464365879074, 0.574306776514277, 0.798747881548479, 0.761049865977839, 0.187582360114902, 0.389573147986084, 0.0845758498180658, 0.193420097930357, 0.929363233037293, 0.429516641655937, 0.446772309718654, 0.847454249858856, 0.110255212988704, 0.774033231893554, 0.201146936044097, 0.62304378952831, 0.348265837877989, 0.809535179985687, 0.894545938586816, 0.630687187891454, 0.665214385371655, 0.733886488247663, 0.257184484275058, 0.780112599954009, 0.817367537878454, 0.287890260107815, 0.787725535687059, 0.191033154726028, 0.117012477712706, 0.338717379141599, 0.0817001643590629, 0.386637721676379, 0.51102749677375, 0.540787796722725, 0.23289854475297, 0.221788419177756, 0.471859542420134, 0.537237155484036, 0.368613599566743, 0.150554252788424, 0.676896716468036, 0.480829717125744, 0.911065152380615, 0.0483708188403398, 0.49158373544924, 0.561625403584912, 0.0815788926556706, 0.751979266060516, 0.723712702980265, 0.325844535604119, 0.869029857218266, 0.892009812640026, 0.336395199177787, 0.719216911122203, 0.955794357694685, 0.656443310435861, 0.259409668389708, 0.683754094643518, 0.828849875601009, 0.355101670371369, 0.846042899414897, 0.892865984933451, 0.695409971056506, 0.260376006364822, 0.268189479829744, 0.841690809931606, 0.610031974501908, 0.47546223225072, 0.683591705514118, 0.593444794649258, 0.780960708390921, 0.143384338589385, 0.326638825237751, 0.2752027916722, 0.290622941683978, 0.787266236962751, 0.86016510380432, 0.277321943314746, 0.274742089211941, 0.325247911270708, 0.654088421259075, 0.307858878746629, 0.879440605407581, 0.659455674234778, 0.570902851177379, 0.0439839465543628, 0.573436802253127, 0.495680507272482, 0.485517704859376, 0.43765461188741, 0.909012835705653, 0.345429989974946, 0.109788392903283, 0.158021469833329, 0.87790318345651, 0.632285283645615, 0.40730351745151, 0.157471909886226, 0.535869860090315, 0.948440980864689, 0.242757764877751, 0.811201358679682, 0.489891401259229, 0.887470417656004, 0.425079074222594, 0.395592312794179, 0.928369111381471, 0.569921540794894, 0.223595704184845, 0.592036690330133, 0.359956070315093, 0.744044342311099, 0.12011048803106, 0.609114556340501, 0.939831240335479, 0.47567960084416, 0.848364276345819, 0.445198935922235, 0.582840112736449, 0.83986938581802, 0.937040010234341, 0.895087281242013, 0.786824208451435, 0.106814443366602, 0.303279736777768, 0.493167340522632, 0.5141634515021, 0.952706096461043, 0.788741537835449, 0.171769586857408, 0.261479985434562, 0.104534079553559, 0.301970061380416, 0.812853845767677, 0.668835047399625, 0.0442299146670848, 0.59660063078627, 0.544202910037711, 0.806963152484968, 0.802757356548682, 0.366263069910929, 0.604173335246742, 0.309765885816887, 0.00784141686744988, 0.140549512580037, 0.652929397299886, 0.601682321168482, 0.435287537518889, 0.356610223883763, 0.970653360243887, 0.731727521400899, 0.965512632625178, 0.915235583903268, 0.363481069449335, 0.999033950036392, 0.971530063310638, 0.427387833595276, 0.219182581175119, 0.89479547413066, 0.888806564034894, 0.993078701430932, 0.0337560758925974, 0.831710944417864, 0.178766110679135, 0.830684259301051, 0.506379540078342, 0.579982160590589, 0.170513777993619, 0.65689178626053, 0.793018362717703, 0.0498002802487463, 0.919376697624102, 0.310970374150202, 0.310101188253611, 0.923781271325424, 0.427758785663173, 0.787831901572645, 0.100475390907377, 0.160902523435652, 0.618849675636739, 0.236374046420678, 0.480606143828481, 0.0480671762488782, 0.382023887010291, 0.248459788272157, 0.41569821885787, 0.328813061350957, 0.693738142261282, 0.0913732724729925, 0.057139910524711, 0.922049350570887, 0.773637649836019, 0.757434646831825, 0.150323193287477, 0.42673652805388, 0.535865859827027, 0.104724589502439, 0.670552924508229, 0.109698558226228, 0.856910568894818, 0.109981609508395, 0.0704593833070248, 0.777032404905185, 0.965790145099163, 0.783476631622761, 0.269026844762266, 0.459091481985524, 0.711974095320329, 0.26058131409809, 0.106574735371396, 0.104136201553047, 0.348117026966065, 0.349080126732588, 0.725573307834566, 0.348933273460716, 0.440287645207718, 0.95139475306496, 0.614358824212104, 0.114586313487962, 0.499641386559233, 0.00853660306893289, 0.8615105538629, 0.818602079758421, 0.163421607576311, 0.440141993341967, 0.438089990522712, 0.679660773137584, 0.435637996997684, 0.856740680523217, 0.890071020228788, 0.186280357884243, 0.477642216952518, 0.330806351965293, 0.0138899672310799, 0.480616943212226, 0.140132981818169, 0.725246994290501, 0.813475401140749, 0.405700080795214, 0.388651900459081, 0.446844443678856, 0.623812108067796, 0.00115519040264189, 0.521651360671967, 0.257509704912081, 0.416268890257925, 0.63645083992742, 0.834683788241819, 0.908979180734605, 0.534536072285846, 0.243564166827127, 0.919417274650186, 0.728399001061916, 0.233814248582348, 0.105182233732194, 0.478869322687387, 0.203245147131383, 0.55340980598703, 0.449643652420491, 0.140958508709446, 0.375270870281383, 0.236998088657856, 0.484713553683832, 0.558300890494138, 0.18853874062188, 0.820230488665402, 0.225428071571514, 0.983054645359516, 0.0997209211345762, 0.541257738368586, 0.118799107847735, 0.839147280436009, 0.0395485667977482, 0.687656339257956, 0.263995185028762, 0.625228304415941, 0.23811047617346, 0.0764929577708244, 0.76289638876915, 0.119508393341675, 0.879318092251197, 0.709086898015812, 0.262095224112272, 0.103178451536223, 0.900209273211658, 0.674916695337743, 0.0508597816806287, 0.923291044542566, 0.984620639588684, 0.142677192110568, 0.243976219557226, 0.511807138333097, 0.0191792116966099, 0.295911719091237, 0.493552081286907, 0.981733283028007, 0.817514467518777, 0.11783041106537, 0.0705767578911036, 0.933777756756172, 0.773820344358683, 0.607554846443236, 0.15139356884174, 0.689231060678139, 0.861015879549086, 0.425067678792402, 0.531470088055357, 0.687517110258341, 0.255856362404302, 0.123642346123233, 0.600707243429497, 0.359583769692108, 0.677518357057124, 0.231058179168031, 0.637597231427208, 0.0887420277576894, 0.351495313923806, 0.184715630020946, 0.168845598120242, 0.777236323803663, 0.105660684639588, 0.342579505173489, 0.616032951511443, 0.942403003107756, 0.0443276532460004, 0.205175886163488, 0.478292324347422, 0.860941132996231, 0.199609839823097, 0.732164270244539, 0.562593021895736, 0.00427875551395118, 0.868976009311154, 0.187780813779682, 0.356872584205121, 0.668426118092611, 0.0684287492185831, 0.358238467946649, 0.046833946602419, 0.558172404067591, 0.365204871399328, 0.214712966699153, 0.763529076008126, 0.332580101210624, 0.232751482399181, 0.0823665021453053, 0.36460575577803, 0.427696462254971, 0.69724699226208, 0.360589698888361, 0.11683858348988, 0.500445061130449, 0.494564316002652, 0.798774332273751, 0.549247197806835, 0.848529055947438, 0.245508210035041, 0.97398593230173, 0.838692541932687, 0.130799028556794, 0.963563247118145, 0.588017177768052, 0.200867828447372, 0.991661249892786, 0.0484734447672963, 0.452220614766702, 0.920244762441143, 0.536236512241885, 0.911240951390937, 0.563780016731471, 0.90473459311761, 0.771885568276048, 0.773198749870062, 0.695104836719111, 0.53555715107359, 0.60664995200932, 0.923804142512381, 0.985135067487136, 0.386300321435556, 0.615385624347255, 0.261513020377606, 0.780387516831979, 0.330964338732883, 0.662278185598552, 0.143755208468065, 0.871003516949713, 0.788686300162226, 0.0173899095971137, 0.254608506802469, 0.407629959052429, 0.957350495504215, 0.420931679196656, 0.37847066950053, 0.290935346856713, 0.659714918816462, 0.921192525653169, 0.768788252491504, 0.689474195707589, 0.707303153350949, 0.927519556134939, 0.601155918091536, 0.00856308010406792, 0.322611405747011, 0.748408074257895, 0.474707173882052, 0.341104252263904, 0.508191651897505, 0.918900255346671, 0.0543964433018118, 0.234711023047566, 0.552779859630391, 0.794247420271859, 0.242873801384121, 0.895189871778712, 0.813295898493379, 0.745297632180154, 0.778511429205537, 0.467708717333153, 0.61307783680968, 0.0393494954332709, 0.383770730812103, 0.442216142546386, 0.0291177628096193, 0.868863707175478, 0.799872473115101, 0.668088880833238, 0.318278557620943, 0.810965090757236, 0.272548764245585, 0.834638703148812, 0.210673971334472, 0.209497225936502, 0.70577805233188, 0.071397635852918, 0.00441402522847056, 0.600986445788294, 0.165233140345663, 0.817520638694987, 0.590692168101668, 0.135569302598014, 0.190429284237325, 0.29033498140052, 0.464987769722939, 0.564331717323512, 0.918000064790249, 0.390956711489707, 0.0297598470933735, 0.764349983772263, 0.878523494582623, 0.387537829810753, 0.691174965118989, 0.503392387647182, 0.450055067660287, 0.898663058876991, 0.685648529324681, 0.905405391007662, 0.45847934181802, 0.166058250702918, 0.533124402631074, 0.370214612223208, 0.669286820106208, 0.316834759432822, 0.457845675759017, 0.249028414720669, 0.521584699628875, 0.751383719965816, 0.724630284588784, 0.0472686244174838, 0.993456530384719, 0.00453077722340822, 0.893412315286696, 0.965319858165458, 0.644843640737236, 0.150848989840597, 0.216302734334022, 0.757538170553744, 0.409974431619048, 0.290042056236416, 0.872164684580639, 0.161378763616085, 0.164533198811114, 0.952619538176805, 0.518712417688221, 0.133064034860581, 0.932216360932216, 0.168749639764428, 0.407925073057413, 0.0622985155787319, 0.326416140655056, 0.367357443086803, 0.419235802488402, 0.77398494258523, 0.523165298392996, 0.585780915338546, 0.553169919643551, 0.768025022232905, 0.282680637435988, 0.550504053011537, 0.505135010229424, 0.658654438564554, 0.99809442833066, 0.757610156666487, 0.267130474559963, 0.0177874492947012, 0.366808643564582, 0.363433910999447, 0.0276958507020026, 0.777238436043262, 0.15698992786929, 0.557015623664483, 0.764903886243701, 0.665829723933712, 0.0637720166705549, 0.747527814935893, 0.757861700840294, 0.751778526464477, 0.825773499440402, 0.0269514804240316, 0.601419975981116, 0.410736301448196, 0.187344479840249, 0.0809977815952152, 0.732995708240196, 0.241526901954785, 0.447395421564579, 0.129500278504565, 0.586924872826785, 0.293827730696648, 0.739933574805036, 0.244407914113253, 0.423606159398332, 0.744285789551213, 0.406977724749595, 0.327198236715049, 0.263818956911564, 0.835687031270936, 0.6583665413782, 0.745569326682016, 0.253629117738456, 0.075746632181108, 0.698585675330833, 0.734846703475341, 0.0794395788107067, 0.297597653232515, 0.226945955539122, 0.0887358929030597, 0.593782047042623, 0.25122709502466, 0.921457767253742, 0.486045261379331, 0.703115621581674, 0.0411302279680967, 0.832266993820667, 0.8838588767685, 0.982112415367737, 0.611516975332052, 0.977422676282004, 0.15084388339892, 0.476870952174067, 0.889667925657704, 0.167479205643758, 0.194147676695138, 0.551918046781793, 0.954744165530428, 0.0979792855214328, 0.913188366452232, 0.51368887652643, 0.359532282222062, 0.703484060941264, 0.750594367040321, 0.884380854899064, 0.565257045207545, 0.0518994862213731, 0.492342258105054, 0.268469776026905, 0.659017743775621, 0.056292116176337, 0.753069396130741, 0.0717194359749556, 0.367306926520541, 0.0162636949680746, 0.923359752632678, 0.688099758001044, 0.173753711394966, 0.639880902832374, 0.893638260429725, 0.945135903777555, 0.175209388369694, 0.673022252274677, 0.815139890415594, 0.275764604797587, 0.18579624267295, 0.555274309124798, 0.846777859143913, 0.593552528880537, 0.249538889154792, 0.720123847713694, 0.597237782087177, 0.43111270130612, 0.378759125014767, 0.930601840373129, 0.0766709980089217, 0.906072779325768, 0.385907034622505, 0.943809809861705, 0.455914427293465, 0.542853356106207, 0.672377650626004, 0.391363677801564, 0.92288251942955, 0.122645775787532, 0.362971076741815, 0.365028861444443, 0.238366901874542, 0.123455196153373, 0.311968124005944, 0.351568781072274, 0.0820802187081426, 0.501918642316014, 0.302601850824431, 0.175752124981955, 0.46662265365012, 0.464865398826078, 0.018744662636891, 0.650246721692383, 0.573831059969962, 0.215232956688851, 0.855521821184084, 0.0429143172223121, 0.812330512097105, 0.920140116475523, 0.768100046785548, 0.593842932488769, 0.480555332498625, 0.571883235126734, 0.101228573359549, 0.793277537683025, 0.226387941511348, 0.938330616569147, 0.84577322867699, 0.932018195511773, 0.831407999852672, 0.91638377122581, 0.936711949761957, 0.0444176474120468, 0.514003484044224, 0.30594765371643, 0.104984176578, 0.804522185353562, 0.565262526506558, 0.984585514757782, 0.78573657781817, 0.42003745236434, 0.064706644974649, 0.383989221416414, 0.159580703824759, 0.166109792189673, 0.395684700924903, 0.707049048738554, 0.376608909573406, 0.759392490843311, 0.8300942836795, 0.186839784728363, 0.481364018516615, 0.246261143591255, 0.977377955336124, 0.36998642468825, 0.269429496955127, 0.457328602671623, 0.208083829842508, 0.142531008925289, 0.603367624571547, 0.980121567146853, 0.229691544780508, 0.788832661695778, 0.266795797972009, 0.380254643969238, 0.208204367896542, 0.536905775312334, 0.654469801811501, 0.244600749807432, 0.0290566049516201, 0.723483355017379, 0.71507354080677, 0.960700026946142, 0.236628360347822, 0.737706815358251, 0.197007834212855, 0.783920046407729, 0.685308757005259, 0.743580226087943, 0.393913518637419, 0.226052533136681, 0.696723981527612, 0.0855445447377861, 0.467781737912446, 0.62662582565099, 0.415112598566338, 0.442008815705776, 0.747535589616746, 0.379840773297474, 0.766632754122838, 0.963479192228988, 0.119482886744663, 0.757011813344434, 0.791617208858952, 0.921725121792406, 0.0284186541102827, 0.517569489078596, 0.294609722215682, 0.481481596361846, 0.814313305076212, 0.110453994711861, 0.52747038169764, 0.528971376596019, 0.794703045859933, 0.666730070486665, 0.681678359629586, 0.220528309233487, 0.734775245189667, 0.93384270905517, 0.669754810165614, 0.669288324657828, 0.640468165278435, 0.957749698543921, 0.980199143756181, 0.998060603393242, 0.964922613231465, 0.779460688587278, 0.956471039215103, 0.350016503827646, 0.0997987152077258, 0.237352626165375, 0.0973105805460364, 0.0747734464239329, 0.632213044445962, 0.185457832645625, 0.934880964225158, 0.918313729576766, 0.18923007324338, 0.557396799558774, 0.143888465128839, 0.927334373816848, 0.531134848948568, 0.893827183870599, 0.691248961957172, 0.249422418884933, 0.138440966838971, 0.637951881159097, 0.805996283655986, 0.018805745523423, 0.00536344619467854, 0.666015117894858, 0.350774240447208, 0.17758005252108, 0.882216204889119, 0.0243230126798153, 0.751601004507393, 0.139050867641345, 0.716389204608276, 0.704338033916429, 0.0328483760822564, 0.645069889258593, 0.951618745923042, 0.276807634392753, 0.488380788592622, 0.720545815769583, 0.138575366698205, 0.699340729275718, 0.880144227994606, 0.192365314811468, 0.917198711074889, 0.0965771360788494, 0.440929555566981, 0.891275165602565, 0.312840463826433, 0.968654358526692, 0.430974711896852, 0.517269360134378, 0.462775385240093, 0.872985275229439, 0.142766386969015, 0.609975822968408, 0.374731294577941, 0.047704704105854, 0.564239896833897, 0.779209388652816, 0.472487004240975, 0.317449492402375, 0.551489867502823, 0.398992333095521, 0.570430210325867, 0.0233488958328962, 0.91587196663022, 0.161511453799903), actual = c(1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"