Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.5852945        0.5852945 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.5341376        0.5341376        0.5130969 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.00423485459759831, 0.453581914305687, 0.882115554763004, 0.995827391277999, 0.895093376282603, 0.825621542986482, 0.53638724423945, 0.669660703279078, 0.261479426873848, 0.124459382379428, 0.0675838673487306, 0.705316299106926, 0.657580252969638, 0.466082046972588, 0.935648975195363, 0.718527149409056, 0.593382987193763, 0.821949454490095, 0.789532050956041, 0.37812404660508, 0.204146794974804, 0.474893695442006, 0.963055780623108, 0.590652246261016, 0.632091208128259, 0.286690281936899, 0.193106008693576, 0.185490467119962, 0.761169810779393, 0.967816145624965, 0.0174678440671414, 0.60991480271332, 0.183759927051142, 0.258887458359823, 0.615268183639273, 0.524622582131997, 0.972701112274081, 0.464391203364357, 0.347605097107589, 0.435191896278411, 0.170587635133415, 0.400586399249732, 0.635961221298203, 0.98501880443655, 0.991467258427292, 0.284723833203316, 0.0327180891763419, 0.0795099397655576, 0.0417711052577943, 0.449354543816298, 0.912979539483786, 0.458157747751102, 0.260553678497672, 0.0667621677275747, 0.587024031905457, 0.404601444723085, 0.468506808858365, 0.384611963294446, 0.268476604949683, 0.428461295552552, 0.0188627589959651, 0.403809362556785, 0.915690835332498, 0.415541043737903, 0.263835006160662, 0.246512450277805, 0.091060099657625, 0.456584413768724, 0.90262092417106, 0.913478646893054, 0.379985658451915, 0.0724075234029442, 0.445002838037908, 0.45722936000675, 0.220965933986008, 0.283818873343989, 0.125497399363667, 0.413803388131782, 0.79944932484068, 0.398545634467155, 0.967891493346542, 0.0915166351478547, 0.318147190846503, 0.124887121841311, 0.917475688969716, 0.306913816137239, 0.829301448073238, 0.705680324928835, 0.58140994864516, 0.339616043260321, 0.84958323976025, 0.405939404619858, 0.161450568353757, 0.355243568075821, 0.835333554306999, 0.906005096854642, 0.320061304373667, 0.0193041916936636, 0.549544658511877, 0.656072176527232, 0.97525952802971, 0.708706898381934, 0.15827243286185, 0.818740632152185, 0.631575586739928, 0.833159519825131, 0.749335641507059, 0.66893107816577, 0.80479512643069, 0.947577208979055, 0.559377832105383, 0.614237393718213, 0.498827456263825, 0.78141730488278, 0.454729550518095, 0.497740080347285, 0.704760666238144, 0.809349341085181, 0.610501631861553, 0.0166654668282717, 0.389674467267469, 0.50903058773838, 0.471364783588797, 0.725696188630536, 0.0255964365787804, 0.247697291430086, 0.558049792423844, 0.554389795288444, 0.20360244740732, 0.932250148383901, 0.52344117523171, 0.367072796449065, 0.16659161192365, 0.681107980664819, 0.813786645419896, 0.313835385488346, 0.385988890891895, 0.578313805628568, 0.404987864429131, 0.855855254922062, 0.203729152679443, 0.48838368197903, 0.457960547646508, 0.700874706497416, 0.19497606344521, 0.701548071345314, 0.48438808391802, 0.698695939034224, 0.0683044521138072, 0.972967640729621, 0.911656479816884, 0.773874299135059, 0.963592669460922, 0.708126385463402, 0.834276580251753, 0.225087585393339, 0.395526674110442, 0.387136161793023, 0.0163378645665944, 0.00994710181839764, 0.796172173228115, 0.430654040537775, 0.303391637280583, 0.103462101193145, 0.380151903489605, 0.469107771292329, 0.554213004186749, 0.959192042471841, 0.819918056484312, 0.0800173787865788, 0.773111323360354, 0.00873845629394054, 0.25059312209487, 0.61804002057761, 0.416045820573345, 0.12371307448484, 0.192712429212406, 0.0672266543842852, 0.395671232836321, 0.0917550909798592, 0.397076170891523, 0.119385732803494, 0.902806254802272, 0.700764873297885, 0.913048559101298, 0.148940165759996, 0.0608036508783698, 0.9701185033191, 0.0722513124346733, 0.286181087838486, 0.677993621909991, 0.359431435354054, 0.0112058646045625, 0.808585682185367, 0.236444214824587, 0.640718612354249, 0.504338702419773, 0.648691966198385, 0.95953627047129, 0.111604869598523, 0.529801460448653, 0.240924020763487, 0.292129466775805, 0.100254756398499, 0.544142414117232, 0.0327516063116491, 0.0829287921078503, 0.826576533494517, 0.20773166581057, 0.876679554581642, 0.0254988232627511, 0.745318979024887, 0.058511472074315, 0.508316085906699, 0.365637214621529, 0.993201478617266, 0.82696967269294, 0.914009447675198, 0.351968077244237, 0.2468132507056, 0.36673834361136, 0.0248366792220622, 0.552167059620842, 0.35656705009751, 0.563598633976653, 0.0316100281197578, 0.508795945672318, 0.548411697382107, 0.0125973748508841, 0.961133734323084, 0.224138499703258, 0.300044847652316, 0.862518261419609, 0.346035579685122, 0.815058609936386, 0.346446153940633, 0.344124815659598, 0.441553911427036, 0.534208446741104, 0.209155378863215, 0.220893546938896, 0.932599435327575, 0.958338592667133, 0.7817242750898, 0.361303184647113, 0.870767302345484, 0.24518872750923, 0.0345667523797601, 0.431341473944485, 0.0517901896964759, 0.806733452482149, 0.886662253411487, 0.980224687838927, 0.639357963576913, 0.228144973050803, 0.307555152568966, 0.450700727757066, 0.590271590044722, 0.161452285945415, 0.768467797897756, 0.542629395145923, 0.575826492393389, 0.853639969835058, 0.532922059297562, 0.545918880961835, 0.242721747141331, 0.6889615340624, 0.59654338285327, 0.673834536224604, 0.457779058488086, 0.251978868851438, 0.358357395511121, 0.210197728592902, 0.9772197608836, 0.412068426609039, 0.571725090267137, 0.436096543446183, 0.529090632451698, 0.361628199694678, 0.867341473000124, 0.536164591554552, 0.796059461310506, 0.171679065097123, 0.189478827407584, 0.261101745767519, 0.65234684292227, 0.0436110349837691, 0.880816492252052, 0.610994067043066, 0.961686954600736, 0.118808711180463, 0.102878704434261, 0.888253188924864, 0.108915134565905, 0.255798250902444, 0.0579893293324858, 0.416179471416399, 0.452811910072342, 0.608355275122449, 0.749703387264162, 0.656780733261257, 0.0429162534419447, 0.329856155440211, 0.682435269234702, 0.428671451052651, 0.728416323428974, 0.146846025483683, 0.559496665606275, 0.212786620249972, 0.118558024754748, 0.516615023370832, 0.360691861016676, 0.979330296395347, 0.152471977751702, 0.543177306419238, 0.514884047908708, 0.486414789687842, 0.448799685342237, 0.661273862002417, 0.0559059719089419, 0.482171416515484, 0.399128848221153, 0.970524485921487, 0.607021908974275, 0.995739521691576, 0.671593129402027, 0.787900291150436, 0.389257375849411, 0.214148909319192, 0.385476377326995, 0.403171204030514, 0.0474043455906212, 0.445414289832115, 0.278229327872396, 0.797259159153327, 0.491257605608553, 0.205228440463543, 0.973962645977736, 0.565660669701174, 0.906572411069646, 0.378275808412582, 0.965201590908691, 0.661663620732725, 0.901038778712973, 0.654935903148726, 0.0225536783691496, 0.667212633416057, 0.911657598102465, 0.710121731506661, 0.274829290108755, 0.426015623379499, 0.74843149445951, 0.448249546112493, 0.225975462934002, 0.528289990499616, 0.579216775251552, 0.900138984667137, 0.689015232026577, 0.298500030068681, 0.535669677890837, 0.530230357078835, 0.649163658032194, 0.44943002006039, 0.710098628187552, 0.881754144327715, 0.471244820160791, 0.679033956956118, 0.704843471059576, 0.819799610879272, 0.755254351301119, 0.585704287979752, 0.133092851843685, 0.15282384888269, 0.866339802974835, 0.242492274846882, 0.983591199386865, 0.913503072224557, 0.571640735724941, 0.662724304245785, 0.512097244616598, 0.198013564106077, 0.770347430603579, 0.835505535360426, 0.172000533435494, 0.269865006906912, 0.0817218034062535, 0.888236246304587, 0.771437693852931, 0.634733546990901, 0.640677773859352, 0.642751262290403, 0.642509448342025, 0.126912103965878, 0.328226354438812, 0.549829044146463, 0.61635962035507, 0.898917237762362, 0.650661593303084, 0.164921301649883, 0.178678382188082, 0.331488034920767, 0.243028429802507, 0.929795870091766, 0.0175464141648263, 0.0705279563553631, 0.632739799795672, 0.712739302543923, 0.467964370269328, 0.42351608001627, 0.783809277229011, 0.539887390332296, 0.227544893976301, 0.302121386164799, 0.799737002002075, 0.229508249787614, 0.95445438940078, 0.565725683700293, 0.672673381399363, 0.927324635908008, 0.90175634296611, 0.584310648962855, 0.198859357042238, 0.0689991870895028, 0.899979699635878, 0.629738986026496, 0.337873964570463, 0.550368228927255, 0.74336445867084, 0.348286025226116, 0.266057396307588, 0.95836602896452, 0.468582820147276, 0.555680062854663, 0.732900794362649, 0.591715309768915, 0.932033428456634, 0.574555992847309, 0.0123420788440853, 0.678495382890105, 0.730693294433877, 0.435278843156993, 0.918099060887471, 0.61145758908242, 0.759809482144192, 0.659745843382552, 0.329097792273387, 0.934107709908858, 0.7400266174227, 0.884014258859679, 0.965562844416127, 0.777520450297743, 0.405416501685977, 0.397798395715654, 0.906202966812998, 0.656437830533832, 0.796516547678038, 0.163753631990403, 0.948462153086439, 0.592493788572028, 0.184926334302872, 0.469145563896745, 0.70323600480333, 0.10449581942521, 0.700070310849696, 0.785782750928774, 0.229820830281824, 0.472996067721397, 0.659116107271984, 0.0550111352931708, 0.348198847379535, 0.942768798209727, 0.276514118304476, 0.164621157804504, 0.540058269165456, 0.962783986935392, 0.442221317207441, 0.569607611512765, 0.226949215633795, 0.350765691138804, 0.851503635756671, 0.14853313472122, 0.444871587911621, 0.331371679902077, 0.235207235207781, 0.0393026920501143, 0.999767130240798, 0.861482811393216, 0.756895502330735, 0.709723273525015, 0.93474509078078, 0.222810972249135, 0.746664508478716, 0.961782387457788, 0.388258087914437, 0.49626404652372, 0.000265876296907663, 0.479693236527964, 0.428943205624819, 0.33676383132115, 0.624866165453568, 0.558633198961616, 0.850366326747462, 0.273917263839394, 0.306206136709079, 0.566627277992666, 0.293186561670154, 0.807298693573102, 0.31915700645186, 0.381474415073171, 0.59669293044135, 0.260977101512253, 0.739889397053048, 0.892784512368962, 0.256949601462111, 0.0830058641731739, 0.00421379203908145, 0.208705481374636, 0.741492165252566, 0.163753709523007, 0.815104962093756, 0.209676982369274, 0.826286722207442, 0.975052807712927, 0.842141260160133, 0.655096059432253, 0.195456240326166, 0.624656788771972, 0.564201399451122, 0.590564345009625, 0.521086212014779, 0.150381413288414, 0.751923900097609, 0.74998061126098, 0.285902956500649, 0.434514932101592, 0.346889999927953, 0.365539027377963, 0.722654097247869, 0.442034577019513, 0.833370763109997, 0.28404216747731, 0.821611882885918, 0.561684943269938, 0.663908860879019, 0.392917986726388, 0.719201540341601, 0.516898873727769, 0.866454647853971, 0.679618098773062, 0.46778033208102, 0.172716360306367, 0.311772951390594, 0.778637103736401, 0.767292906297371, 0.643530853325501, 0.50368217122741, 0.41442085034214, 0.0863736951723695, 0.601824567420408, 0.0967544030863792, 0.145345904864371, 0.250067722285166, 0.673953485675156, 0.966912467265502, 0.617645314428955, 0.417454350506887, 0.917483452009037, 0.0876879915595055, 0.313134636031464, 0.447441203985363, 0.619416114175692, 0.432307538809255, 0.163389584748074, 0.196524906205013, 0.130021838471293, 0.848071398911998, 0.896546758711338, 0.22219579690136, 0.0226269064005464, 0.706833597505465, 0.512493928428739, 0.329936005640775, 0.658220210112631, 0.297373980982229, 0.792450058273971, 0.677575347712263, 0.37681857845746, 0.202083440497518, 0.970016962382942, 0.0349093209952116, 0.548740037018433, 0.274585665669292, 0.991251034429297, 0.601749621564522, 0.99500727490522, 0.524433141574264, 0.75569737656042, 0.606970127206296, 0.20515787624754, 0.320259559666738, 0.0585291360039264, 0.616942951688543, 0.518094665370882, 0.559334255754948, 0.996230192715302, 0.227101230528206, 0.891605915967375, 0.0480807079002261, 0.205136124277487, 0.891595674213022, 0.442410331685096, 0.747235762653872, 0.114949401235208, 0.443206671392545, 0.339297727914527, 0.317243909696117, 0.613324053119868, 0.736309129279107, 0.730634076520801, 0.72105033765547, 0.371191285783425, 0.182999786455184, 0.00221842899918556, 0.300337522756308, 0.392025008797646, 0.755949230864644, 0.647772878641263, 0.462944303173572, 0.0631248282734305, 0.359979247907177, 0.500452722422779, 0.219712121644989, 0.507380218943581, 0.961307112593204, 0.98441192577593, 0.84352641296573, 0.457536209607497, 0.631924652028829, 0.761781672015786, 0.160177003126591, 0.44344472209923, 0.816175959073007, 0.221248045098037, 0.530818254454061, 0.0704669314436615, 0.867337679024786, 0.547921395627782, 0.705625793663785, 0.909606512170285, 0.0910028675571084, 0.249484643572941, 0.240957446396351, 0.210769708268344, 0.423133188392967, 0.45707887224853, 0.829273188253865, 0.349451114423573, 0.288524530827999, 0.375402628211305, 0.968907242175192, 0.994974982691929, 0.128408076474443, 0.688439509598538, 0.124914539279416, 0.503997230203822, 0.5149558451958, 0.00995939993299544, 0.00940806861035526, 0.24168767221272, 0.288702704245225, 0.777519334806129, 0.122545233229175, 0.428658403689042, 0.771638852544129, 0.239781942917034, 0.308923471253365, 0.787060079630464, 0.290888944873586, 0.787819979945198, 0.456874911673367, 0.718360668979585, 0.712079165037721, 0.0798288090154529, 0.0262714051641524, 0.533888259669766, 0.280032541835681, 0.530680031515658, 0.913992704590783, 0.151585672516376, 0.683084561023861, 0.13052865723148, 0.4952928556595, 0.273932276526466, 0.199301563668996, 0.467611687956378, 0.811099377693608, 0.382582652615383, 0.254237673943862, 0.598517399746925, 0.855331494705752, 0.0295879442710429, 0.561567521421239, 0.032339722616598, 0.0694480242673308, 0.104368442669511, 0.164013025583699, 0.506909864023328, 0.0832831477746367, 0.81632765289396, 0.198567199055105, 0.782803222537041, 0.631644968176261, 0.522827168693766, 0.763844451168552, 0.190248324535787, 0.782902476144955, 0.782484060619026, 0.926556426798925, 0.662206872133538, 0.311662615742534, 0.729383714962751, 0.864006710937247, 0.947922733845189, 0.504394315648824, 0.730184534331784, 0.635554736247286, 0.644919417332858, 0.0214826592709869, 0.502725762780756, 0.88350710645318, 0.31452302262187, 0.998517144005746, 0.527665574802086, 0.0562911953311414, 0.357805956620723, 0.592510872287676, 0.691737111192197, 0.273850543191656, 0.83154629683122, 0.466174235101789, 0.998843155568466, 0.228850477607921, 0.761437780456617, 0.184227722464129, 0.626780634513125, 0.138695595785975, 0.390387428691611, 0.25362704647705, 0.977299282792956, 0.26238993415609, 0.912922666175291, 0.544670357368886, 0.131300400011241, 0.847590822493657, 0.0679851935710758, 0.550648079486564, 0.198343820404261, 0.560017368756235, 0.774994698120281, 0.100688628153875, 0.967088720062748, 0.674721135525033, 0.753717865096405, 0.623775018844754, 0.29748861421831, 0.875285250833258, 0.224324422888458, 0.373347235145047, 0.151049183681607, 0.459271045634523, 0.89798096450977, 0.780450692400336, 0.0764097243081778, 0.232830930035561, 0.0303015063982457, 0.611662450712174, 0.881128473440185, 0.876813485985622, 0.408516288036481, 0.957698890706524, 0.897922699339688, 0.95375294610858, 0.127967720618472, 0.222965251421556, 0.975666000274941, 0.453316685976461, 0.203294944483787, 0.842141340021044, 0.685741701396182, 0.581352379871532, 0.164736330509186, 0.649233700474724, 0.545181333785877, 0.449533600592986, 0.566710666520521, 0.512619844637811, 0.0515457773581147, 0.782853207085282, 0.798953008837998, 0.077192469034344, 0.371695471229032, 0.840445418842137, 0.551439276896417, 0.499069965677336, 0.469901669071987, 0.136059022974223, 0.0405086809769273, 0.832702062092721, 0.741793252760544, 0.494099085917696, 0.686887393239886, 0.76027196017094, 0.888478333130479, 0.859460514038801, 0.0848507087212056, 0.46917037689127, 0.305757619207725, 0.753812999464571, 0.606393401743844, 0.777615549974144, 0.400631387718022, 0.517148455837741, 0.968852136749774, 0.0113192971330136, 0.658868536818773, 0.5759403987322, 0.94444255111739, 0.801385373575613, 0.444062687223777, 0.482510526431724, 0.0681858973111957, 0.782025188207626, 0.380213506286964, 0.994794697035104, 0.259608282474801, 0.120398975443095, 0.369778140448034, 0.739206201629713, 0.486194438068196, 0.0839799784589559, 0.685635699657723, 0.129355850629508, 0.885995183140039, 0.24902656278573, 0.3790995166637, 0.371274966746569, 0.962598646292463, 0.0447887605987489, 0.105210756417364, 0.0248836798127741, 0.792063100961968, 0.344390497077256, 0.616498335730284, 0.418636474991217, 0.94338273210451, 0.746966878417879, 0.678373259492218, 0.135504856240004, 0.638232037890702, 0.601746549829841, 0.123037952696905, 0.673442110884935, 0.202188109047711, 0.664413611637428, 0.719143023481593, 0.532959080999717, 0.594746974762529, 0.28127655480057, 0.298971494892612, 0.783242475474253, 0.491172648966312, 0.735459087882191, 0.234085878124461, 0.751471890602261, 0.873807926196605, 0.193489554338157, 0.371089654043317, 0.508523139636964, 0.333826205693185, 0.184523465344682, 0.590904560405761, 0.633358114166185, 0.216181389987469, 0.170510365860537, 0.726966292597353, 0.610326139023528, 0.534780079731718, 0.54224952426739, 0.422697477042675, 0.83771832077764, 0.00639785174280405, 0.17953516729176, 0.0834803069010377, 0.415435066679493, 0.191705950535834, 0.787881371332332, 0.805782112758607, 0.626672576181591, 0.215473094256595, 0.345652375603095, 0.743894145591184, 0.694592786952853, 0.74996533500962, 0.252173721324652, 0.857369770761579, 0.98055771109648, 0.220210268162191, 0.114633759483695, 0.10287735844031, 0.307571445591748, 0.478167926659808, 0.829528283094987, 0.779028112068772, 0.800807448802516, 0.196049757301807, 0.120277432724833, 0.0882953405380249, 0.17100947862491, 0.218422690872103, 0.293988795951009, 0.0258687986060977, 0.386649953899905, 0.255591552471742, 0.926112259970978, 0.334645466413349, 0.645170853473246, 0.458891999907792, 0.554602023679763, 0.238252558745444, 0.413047714158893, 0.613988928729668, 0.203928710194305, 0.402101126965135, 0.107996892882511, 0.217380242887884, 0.661696274066344, 0.919836813351139, 0.188902413472533, 0.114611292490736, 0.43701115087606, 0.565024252748117, 0.957222584169358, 0.920683802338317, 0.276106238598004, 0.619916323572397, 0.260992025723681, 0.898508030455559, 0.136953390901908, 0.0740117882378399, 0.31252210191451, 0.635734483134001, 0.306077156448737, 0.706866615219042, 0.422410638537258, 0.28962712478824, 0.893040006048977, 0.698918896261603, 0.955537258181721, 0.904392155585811, 0.685761622386053, 0.419393374817446, 0.123659942299128, 0.985241046408191, 0.969362455187365, 0.616581001551822, 0.0799547526985407, 0.631826453376561, 0.365725228562951, 0.220762114739046, 0.806528664426878, 0.881064295535907, 0.772632218431681, 0.10100680636242, 0.571904951706529, 0.757713715778664, 0.788924347609282, 0.845317834988236, 0.460136137669906, 0.847685433691368, 0.253472946817055, 0.608423200435936, 0.548088443465531, 0.257340021198615, 0.676380011253059, 0.516465961467475, 0.868622621055692, 0.800410728203133, 0.508299908135086, 0.836204011924565, 0.751114820595831, 0.358077846700326, 0.0441203445661813, 0.582735086558387, 0.114926313748583, 0.985518262255937, 0.728091749595478, 0.692301204893738, 0.760681384243071, 0.695859110448509, 0.420779640553519), actual = c(1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.869400014868006, 0.360658708028495, 0.212400834076107, 0.91109597729519, 0.198820188641548, 0.272116881329566, 0.165661566657946, 0.962332183727995, 0.00578310457058251, 0.75579901249148, 0.813511319924146, 0.0114710005000234, 0.984326863894239, 0.628328328952193, 0.64003535057418, 0.869489384582266, 0.760367200477049, 0.603049991186708, 0.0414549796842039, 0.869447073200718, 0.702944620046765, 0.312034968752414, 0.640412182314321, 0.490707412362099, 0.527652583085001, 0.976457057055086, 0.995522251585498, 0.640890179900452, 0.00338680576533079, 0.702580515295267, 0.229243810754269, 0.686626502778381, 0.0221993250306696, 0.101881556911394, 0.827246284345165, 0.998544021276757, 0.258621396031231, 0.0152775580063462, 0.00850500911474228, 0.101207702420652, 0.457814408000559, 0.542222751071677, 0.330151265719905, 0.614490676438436, 0.27739146607928, 0.72205273504369, 0.424041853751987, 0.220336437225342, 0.396208772435784, 0.455627225339413, 0.141358933178708, 0.624256668845192, 0.540213521802798, 0.542938922531903, 0.294804177479818, 0.369481487665325, 0.982354476349428, 0.411613109521568, 0.0958531289361417, 0.454754737438634, 0.788336188299581, 0.431829651584849, 0.68821896915324, 0.492119165603071, 0.491512339329347, 0.971428193384781, 0.234876735368744, 0.781874158186838, 0.931846174178645, 0.61811468238011, 0.349355123471469, 0.224047933937982, 0.597897030878812, 0.843817173037678, 0.577941714553162, 0.777052743593231, 0.373080091085285, 0.291843068320304, 0.424800156615674, 0.484426619019359, 0.510881738737226, 0.805044936481863, 0.525305275805295, 0.801767449826002, 0.286457336274907, 0.869176532374695, 0.387153250398114, 0.312216203659773, 0.960431540152058, 0.640901956008747, 0.973969816928729, 0.177033722167835, 0.877647852990776, 0.432335683843121, 0.757112256018445, 0.43209446943365, 0.628674051957205, 0.974381984909996, 0.107254345202819, 0.558041077805683, 0.769508241210133, 0.609780476428568, 0.0282723612617701, 0.777472185203806, 0.908612822880968, 0.15697120805271, 0.553172571584582, 0.0904468479566276, 0.843084723223001, 0.119966425467283, 0.917481167241931, 0.287977972067893, 0.694942873204127, 0.141442616702989, 0.456467472948134, 0.416444327449426, 0.981672134250402, 0.762537762988359, 0.337258270708844, 0.0382590712979436, 0.431652054889128, 0.433337447466329, 0.548034135485068, 0.611606003250927, 0.699391245609149, 0.0723346832673997, 0.613194179954007, 0.924846896668896, 0.531316942069679, 0.211086043156683, 0.920314129674807, 0.581092321313918, 0.79021475976333, 0.668488563038409, 0.472507365047932, 0.687479416374117, 0.576534668216482, 0.804828873369843, 0.893637633882463, 0.728373350109905, 0.982175929006189, 0.473661376163363, 0.581788700073957, 0.317229714943096, 0.7550501502119, 0.475469094468281, 0.660301625728607, 0.985073391580954, 0.499120467342436, 0.857936340384185, 0.740174790844321, 0.870045595569536, 0.510261107003316, 0.0983096829149872, 0.8830549151171, 0.337961866986006, 0.742348310770467, 0.201122866477817, 0.627609031740576, 0.486549156252295, 0.919157622382045, 0.856171601684764, 0.662558548152447, 0.252123210811988, 0.46333100926131, 0.933754192432389, 0.700983389280736, 0.273752456065267, 0.0590690860990435, 0.209926493698731, 0.894471440464258, 0.640185769414529, 0.13030206062831, 0.974576802924275, 0.684972429648042, 0.458393315551803, 0.0832086291629821, 0.755234459182248, 0.384587037144229, 0.0409508016891778, 0.836493150796741, 0.296164968749508, 0.802641900023445, 0.536792417056859, 0.0359187419526279, 0.579394183587283, 0.575453317724168, 0.515225735493004, 0.594790491275489, 0.332117684651166, 0.528355555376038, 0.0277261433657259, 0.232209231471643, 0.975988510064781, 0.881444444414228, 0.370554322842509, 0.569218765944242, 0.976390128489584, 0.0706983262207359, 0.969146860996261, 0.5662066352088, 0.801175455795601, 0.916720111854374, 0.762747638858855, 0.822607364971191, 0.492542622843757, 0.704805974848568, 0.242892114445567, 0.949986361898482, 0.832703933585435, 0.316216906998307, 0.126482219202444, 0.134591485839337, 0.741176159353927, 0.383805732242763, 0.332874357234687, 0.00659333681687713, 0.846681895200163, 0.887978632934391, 0.459632544545457, 0.0287387052085251, 0.857911346014589, 0.239083973225206, 0.932146013947204, 0.121606467291713, 0.944444329012185, 0.579928972991183, 0.897945740725845, 0.00139882764779031, 0.385383001063019, 0.231039453297853, 0.212769946316257, 0.977712438208982, 0.447249229066074, 0.334768938133493, 0.837771255057305, 0.266035902546719, 0.0316454577259719, 0.389618555549532, 0.827488196780905, 0.990563031751662, 0.2317183916457, 0.309117066906765, 0.171681896783412, 0.0637645886745304, 0.10377137712203, 0.592580310534686, 0.834781022742391, 0.582036621402949, 0.847227652091533, 0.524023463251069, 0.689610606990755, 0.216392368311062, 0.45142637938261, 0.925031163962558, 0.463368102908134, 0.972486149752513, 0.751274645561352, 0.0472592876758426, 0.0696081344503909, 0.0541740406770259, 0.102907930267975, 0.839461914962158, 0.762319625820965, 0.237959848949686, 0.0701881255954504, 0.801256701583043, 0.514763220679015, 0.95551027962938, 0.273112220456824, 0.536290626740083, 0.661836066050455, 0.209604769945145, 0.924647710286081, 0.149192786775529, 0.869734584586695, 0.186979222344235, 0.741492062574252, 0.459422516170889, 0.339966517174616, 0.760672813514248, 0.734506729291752, 0.249152150703594, 0.353312705410644, 0.359587151324376, 0.563206534832716, 0.205169573426247, 0.165698894299567, 0.622955893399194, 0.375262866262347, 0.600226938724518, 0.268790758913383, 0.88843945437111, 0.285884212702513, 0.162845585029572, 0.0187053792178631, 0.301591966534033, 0.701845430536196, 0.392804813105613, 0.318836117861792, 0.956977797904983, 0.573077455163002, 0.585641279583797, 0.816687342943624, 0.90498227160424, 0.4120750608854, 0.842066659824923, 0.396462312666699, 0.867825724883005, 0.181811342947185, 0.703058186452836, 0.873162826756015, 0.873155170818791, 0.656595702981576, 0.604087781626731, 0.377585222478956, 0.888571584364399, 0.350157694891095, 0.563021467300132, 0.545707078650594, 0.810378004796803, 0.867830655071884, 0.937853651121259, 0.811527639627457, 0.863687986740842, 0.662337207468227, 0.800085900118575, 0.513030787464231, 0.755704745883122, 0.651379529619589, 0.04890215722844, 0.266502695856616, 0.661077277502045, 0.361883921083063, 0.720334223704413, 0.154669714393094, 0.693617976037785, 0.73738076002337, 0.307373547926545, 0.311849986203015, 0.706804491812363, 0.662592058768496, 0.792027710704133, 0.520299834664911, 0.734434701735154, 0.803343043429777, 0.219196897698566, 0.745360384462401, 0.77429370302707, 0.541999920038506, 0.536225886782631, 0.0829727773088962, 0.552515996154398, 0.743106835288927, 0.617641810560599, 0.200454760110006, 0.691335742361844, 0.472965091699734, 0.649805398890749, 0.658865289529786, 0.381120740436018, 0.152607964351773, 0.431932420702651, 0.158014013664797, 0.987077545374632, 0.292171138571575, 0.929416394326836, 0.998673658818007, 0.862238067900762, 0.664026176091284, 0.777691431343555, 0.651051383232698, 0.875095998169854, 0.666072056861594, 0.584514210466295, 0.895817369688302, 0.483208919875324, 0.814869681606069, 0.860741180134937, 0.780761124333367, 0.356650111498311, 0.619281863095239, 0.629332465585321, 0.423991130199283, 0.956156223779544, 0.704430543351918, 0.804697999265045, 0.0586513550952077, 0.156852553598583, 0.682322894223034, 0.680687937419862, 0.546995101962239, 0.25751365465112, 0.78863861062564, 0.0528261556755751, 0.905896812211722, 0.406992680160329, 0.0256358028855175, 0.155774525599554, 0.912460252642632, 0.426682127872482, 0.680394610855728, 0.445040390593931, 0.162739628227428, 0.938276619184762, 0.734744672197849, 0.0674381365533918, 0.706217529717833, 0.883987540844828, 0.750046454137191, 0.37638430017978, 0.765224103815854, 0.454297941643745, 0.21129562635906, 0.0473045862745494, 0.634395824745297, 0.95249434793368, 0.0281257580500096, 0.235330092953518, 0.962313663447276, 0.584563493262976, 0.297125320648775, 0.0475779881235212, 0.0160625192802399, 0.554615044267848, 0.956100502284244, 0.76134186447598, 0.136382283410057, 0.829693925334141, 0.374989706790075, 0.719285918399692, 0.817479995545, 0.839553562691435, 0.987755816662684, 0.0996573637239635, 0.60348033439368, 0.283848692430183, 0.738326652906835, 0.677194394171238, 0.8276037722826, 0.185764552094042, 0.350066476268694, 0.223304235609248, 0.771991775836796, 0.326375687029213, 0.553821332287043, 0.305106140207499, 0.618432904593647, 0.549504992319271, 0.219620967982337, 0.163464745972306, 0.614756363909692, 0.77687827241607, 0.88208112702705, 0.266399983083829, 0.146953710354865, 0.224891630467027, 0.11145009053871, 0.766226714011282, 0.376410180237144, 0.368888622615486, 0.405675515998155, 0.909708405146375, 0.492464940762147, 0.90007237996906, 0.410077399108559, 0.0605464191175997, 0.339804917573929, 0.035746612586081, 0.869734328472987, 0.377552135381848, 0.577869011787698, 0.281250862637535, 0.0247432866599411, 0.0217968139331788, 0.39483640762046, 0.226690003182739, 0.693465153686702, 0.487395821837708, 0.334346131421626, 0.88504805485718, 0.485390256159008, 0.756402899278328, 0.194772017654032, 0.0604068522807211, 0.0583116803318262, 0.966196788940579, 0.462800226407126, 0.807875333353877, 0.177594624925405, 0.996354322414845, 0.76008982793428, 0.588370550889522, 0.0795241026207805, 0.685161845525727, 0.330042428569868, 0.414576372131705, 0.241472591878846, 0.498167593963444, 0.289007032522932, 0.578343718545511, 0.686722254147753, 0.899276117794216, 0.937641564989462, 0.406824441161007, 0.563393184682354, 0.59328249678947, 0.521278999745846, 0.00734928133897483, 0.606240103486925, 0.705142845632508, 0.082940683234483, 0.272848782362416, 0.774972359882668, 0.708571897586808, 0.597639920655638, 0.23039683746174, 0.718386586988345, 0.641709154471755, 0.382924535078928, 0.271379234502092, 0.125690213637426, 0.968324559973553, 0.658885581651703, 0.46738938242197, 0.231761145638302, 0.893747522728518, 0.87911397125572, 0.458209899021313, 0.96192040736787, 0.555929424008355, 0.0778096094727516, 0.0865265533793718, 0.801162402145565, 0.355689499061555, 0.34507424896583, 0.480087469564751, 0.134900602977723, 0.359792871167883, 0.0259428492281586, 0.755725528812036, 0.481216941494495, 0.3958147931844, 0.102742096176371, 0.186927875503898, 0.537986773299053, 0.713159163016826, 0.11539696669206, 0.646599141648039, 0.38787832017988, 0.780412656953558, 0.138947217725217, 0.435054383473471, 0.202036657836288, 0.10783072328195, 0.420459218323231, 0.690498124808073, 0.641517902491614, 0.805212987819687, 0.394146845676005, 0.385329907992855, 0.274351155851036, 0.434434675145894, 0.175258052069694, 0.926484346855432, 0.723589282250032, 0.160900822142139, 0.725933515001088, 0.527237846981734, 0.594982122303918, 0.673140784725547, 0.94098896952346, 0.797693217871711, 0.227763924980536, 0.160954123595729, 0.865013860398903, 0.585299399215728, 0.202012593159452, 0.786892334232107, 0.478989124065265, 0.955600550631061, 0.751200006343424, 0.344990541925654, 0.284881285624579, 0.0616049582604319, 0.478221534518525, 0.719709610799327, 0.639592838240787, 0.933638175716624, 0.144722326891497, 0.25253874482587, 0.750298695405945, 0.613131375517696, 0.539086181204766, 0.0810062589589506, 0.633976737735793, 0.278143535368145, 0.317294377367944, 0.422247546026483, 0.658284366829321, 0.537973742932081, 0.0587204375769943, 0.0917785840574652, 0.403843855950981, 0.608361994847655, 0.362712518544868, 0.123908192617819, 0.284909622510895, 0.636043256148696, 0.0311780783813447, 0.937099040485919, 0.306644892087206, 0.0873436902184039, 0.716878599720076, 0.047324804821983, 0.19971910957247, 0.361116680083796, 0.385017647407949, 0.0257860568817705, 0.828725879546255, 0.985065994551405, 0.0836859997361898, 0.223397355759516, 0.602032209048048, 0.940034130122513, 0.928568384610116, 0.852464038180187, 0.0209919577464461, 0.450213440693915, 0.562347991624847, 0.0189859601669014, 0.403738768072799, 0.386250832816586, 0.426286661764607, 0.960752079728991, 0.277320745633915, 0.752518110210076, 0.691782548790798, 0.134411856066436, 0.446574973175302, 0.59162373514846, 0.0698214755393565, 0.388901383848861, 0.576137670082971, 0.0375298382714391, 0.0941611367743462, 0.165266120573506, 0.240570468595251, 0.751116138882935, 0.393469571368769, 0.269356732955202, 0.156157771125436, 0.106786516262218, 0.142918250989169, 0.640101664001122, 0.711895614629611, 0.226958381943405, 0.220127676147968, 0.0543255927041173, 0.256318550091237, 0.653901983052492, 0.86300265416503, 0.187051745131612, 0.814239863539115, 0.0400657297577709, 0.682604435132816, 0.39781457465142, 0.817387020913884, 0.0655942978337407, 0.919212490320206, 0.817361514316872, 0.741367207607254, 0.274149790406227, 0.182763995602727, 0.228214287664741, 0.97992025851272, 0.726617692271248, 0.738626206293702, 0.994035992305726, 0.603987022070214, 0.334773565642536, 0.214367609471083, 0.4386410205625, 0.740611822810024, 0.437238908838481, 0.867094144690782, 0.678245345596224, 0.36072967806831, 0.509149395860732, 0.505662105046213, 0.738620577845722, 0.395476221339777, 0.274694009451196, 0.376119782449678, 0.545819189399481, 0.875871796859428, 0.972058628452942, 0.078769926680252, 0.941090552834794, 0.742950865533203, 0.275242312112823, 0.115515364799649, 0.352887759217992, 0.0627546228934079, 0.911226688651368, 0.966877648374066, 0.481198287103325, 0.844987285323441, 0.535751736257225, 0.0290796379558742, 0.196935578715056, 0.738884885562584, 0.839251006022096, 0.869477172615007, 0.250875106081367, 0.734858548734337, 0.455673104384914, 0.691701663890854, 0.20044522290118, 0.753161323722452, 0.623893058393151, 0.761432640952989, 0.100485395174474, 0.839442568598315, 0.72503637871705, 0.623862756183371, 0.602449706988409, 0.308095831656829, 0.575162515044212, 0.729457684094086, 0.290464968886226, 0.158247102983296, 0.237103417981416, 0.882916221627966, 0.398147441912442, 0.195430319523439, 0.994531342061237, 0.932409448549151, 0.253332650056109, 0.996973519446328, 0.142780411988497, 0.131421018391848, 0.482171968556941, 0.692822743905708, 0.341256349114701, 0.634807968512177, 0.548922304529697, 0.351289017824456, 0.925953211495653, 0.288283484755084, 0.787924706237391, 0.732413683552295, 0.892927238484845, 0.0800703675486147, 0.349218101473525, 0.313819165807217, 0.333529374795035, 0.924793836195022, 0.758613971062005, 0.809880068292841, 0.222662658896297, 0.764803342521191, 0.614107618574053, 0.571967761963606, 0.560785186011344, 0.526489231502637, 0.587091977940872, 0.749951963778585, 0.333094114437699, 0.406819253228605, 0.756627346388996, 0.327096524415538, 0.409786492818967, 0.568220788612962, 0.759221897693351, 0.724923580884933, 0.433816420845687, 0.0743049837183207, 0.167976773809642, 0.531985950656235, 0.055338207166642, 0.318232289748266, 0.124934586463496, 0.38676128257066, 0.847428979119286, 0.725060393335298, 0.219956395914778, 0.952071158913895, 0.968836923828349, 0.0358367757871747, 0.213313023094088, 0.51337960944511, 0.263833233388141, 0.211510839639232, 0.0289244232699275, 0.0633854130282998, 0.684276768937707, 0.258218442322686, 0.587876121047884, 0.581360036507249, 0.25995770492591, 0.545971618732437, 0.0226651560515165, 0.934213378233835, 0.187925244448707, 0.407448650104925, 0.167505169054493, 0.0696278412360698, 0.322831443278119, 0.11625931202434, 0.303002673434094, 0.974085886031389, 0.955386948771775, 0.876476706238464, 0.283222178230062, 0.20846345112659, 0.332880679052323, 0.387173096183687, 0.723768323427066, 0.529159890720621, 0.324555618455634, 0.758084411500022, 0.322727397084236, 0.833931932924315, 0.964461687020957, 0.881379848346114, 0.682429377920926, 0.285261672921479, 0.332202821969986, 0.748421035474166, 0.435932166175917, 0.316824565641582, 0.423410798888654, 0.717637418303639, 0.198014847235754, 0.725218682549894, 0.366855256957933, 0.182773817563429, 0.265375306131318, 0.378697992069647, 0.0868970842566341, 0.932398042408749, 0.562231486197561, 0.27511019166559, 0.875090812798589, 0.128422749461606, 0.485821730922908, 0.0392515170387924, 0.839662078302354, 0.555210791528225, 0.978259704308584, 0.346222315449268, 0.0836664200760424, 0.111933985259384, 0.0898534450680017, 0.0582376806996763, 0.715534416493028, 0.48891437985003, 0.475685477955267, 0.994656166760251, 0.395137858111411, 0.965365690412, 0.472495641093701, 0.902875038562343, 0.546005704207346, 0.572066424880177, 0.453235149383545, 0.600492620840669, 0.724250103114173, 0.197306357091293, 0.400495912646875, 0.538498953450471, 0.346089761238545, 0.0772796664386988, 0.279724114108831, 0.896245337789878, 0.466239364584908, 0.0654756780713797, 0.851208453532308, 0.818401845404878, 0.514023727970198, 0.385790215339512, 0.352811637567356, 0.11137728812173, 0.581825476838276, 0.173343652160838, 0.792919893050566, 0.567759196972474, 0.539718321058899, 0.475073237903416, 0.646007072646171, 0.277357505867258, 0.409318280871958, 0.165413623675704, 0.229836183367297, 0.00709836510941386, 0.457841865951195, 0.602450811304152, 0.830959944054484, 0.793517967453226, 0.569838614901528, 0.278176720952615, 0.593843853101134, 0.86589309759438, 0.505397033412009, 0.50480449735187, 0.790386426728219, 0.670111582847312, 0.43601135793142, 0.881709568202496, 0.880250716814771, 0.281587171368301, 0.151127310236916, 0.631605018395931, 0.738341775489971, 0.544296805514023, 0.474721332313493, 0.902978736208752, 0.702007120242342, 0.0406829458661377, 0.916708240285516, 0.839085171930492, 0.476108873728663, 0.997560603311285, 0.00515463831834495, 0.944024249911308, 0.310952044324949, 0.480243216035888, 0.568684889469296, 0.199333941563964, 0.653181632515043, 0.132402169983834, 0.0946586464997381, 0.00888754241168499, 0.414257168537006, 0.455814046086743, 0.174453250365332, 0.12831731652841, 0.214019752806053, 0.409396840259433, 0.693030622554943, 0.93012537015602, 0.505360452458262, 0.557267850032076, 0.352139433147386, 0.622518979245797, 0.248605854110792, 0.65227099461481, 0.272880415664986, 0.125646169064566, 0.76997942943126, 0.34645874099806, 0.423262273194268, 0.176071897149086, 0.618854781147093, 0.108211571117863, 0.527574482141063, 0.478347757365555, 0.231243286514655, 0.830968064721674, 0.174913371913135, 0.474724194500595, 0.441000634571537, 0.651459648273885, 0.790228304918855, 0.267382254824042, 0.85667243343778, 0.484813504852355, 0.238782669650391, 0.97991399327293, 0.902872072765604, 0.627214542590082, 0.353802273282781, 0.321375280618668, 0.449533296748996, 0.407289513852447, 0.346983386669308, 0.333634005859494, 0.289523917715997, 0.610123700229451, 0.153744025155902, 0.143225815147161, 0.970711182104424, 0.475045989733189, 0.467702387832105, 0.536909802816808, 0.186978587182239, 0.798183611826971, 0.14222471206449, 0.597517260117456, 0.325700586428866, 0.445156459929422, 0.916609997395426, 0.545961377443746, 0.034052956616506, 0.572210260899737, 0.690789055544883, 0.227033700561151, 0.308538837591186, 0.25250505306758), actual = c(0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"