Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

vector of predicted probabilities

actual

vector of actual outcomes

nmb

named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

list

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.1807531        0.4772298 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4772298        0.4772298        0.4979729 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.116463645827025, 0.43294667545706, 0.837829442461953, 0.770966362673789, 0.0637888601049781, 0.0175068986136466, 0.955561527749524, 0.480480700964108, 0.946515763411298, 0.626004989491776, 0.586568998871371, 0.37782747647725, 0.142537566134706, 0.287636677967384, 0.49257097626105, 0.660285674035549, 0.612154712900519, 0.169432775815949, 0.363453626865521, 0.131349830655381, 0.187638908158988, 0.977951910346746, 0.880618602735922, 0.0586134793702513, 0.953003449598327, 0.304017371730879, 0.916747659444809, 0.0345204009208828, 0.367749884957448, 0.709155978169292, 0.407881394261494, 0.34223189833574, 0.414342106087133, 0.462302763247862, 0.557822131086141, 0.0635934453457594, 0.462289626244456, 0.00714381993748248, 0.869719999143854, 0.736487673828378, 0.285727258073166, 0.0495664251502603, 0.13472111732699, 0.411518539767712, 0.523015804123133, 0.73273524385877, 0.262557653943077, 0.721943790093064, 0.451122682075948, 0.288659637561068, 0.984813543735072, 0.355205500964075, 0.829729145858437, 0.192576815839857, 0.115773194236681, 0.234527223976329, 0.0298229972831905, 0.884084248682484, 0.621339780976996, 0.169340403517708, 0.695961436955258, 0.107620930299163, 0.966658682562411, 0.572849506745115, 0.499266406754032, 0.478271579137072, 0.481026522815228, 0.925249089021236, 0.815766355954111, 0.857705274131149, 0.165243979776278, 0.185341276228428, 0.737210663268343, 0.0156126522924751, 0.511819946579635, 0.839759432477877, 0.130748567404225, 0.189249605173245, 0.354670885484666, 0.0838235432747751, 0.597672206349671, 0.866506838006899, 0.0133814362343401, 0.833344689570367, 0.378529743058607, 0.712207161355764, 0.227980097522959, 0.532703925855458, 0.160114294383675, 0.189050053944811, 0.162927592173219, 0.639881681185216, 0.746708094142377, 0.114037967752665, 0.0842994798440486, 0.951088844565675, 0.428785965079442, 0.0223524055909365, 0.761350983055308, 0.613381471252069, 0.292963386280462, 0.864584808703512, 0.990775026613846, 0.786480941344053, 0.194932281505316, 0.712862864602357, 0.0341873341239989, 0.873556382255629, 0.447542405687273, 0.398141507524997, 0.458194480277598, 0.141316994093359, 0.317632410908118, 0.161045443965122, 0.00481194327585399, 0.438136229291558, 0.869597776560113, 0.108903298852965, 0.973709989571944, 0.20106366276741, 0.0982959028333426, 0.542809372767806, 0.352836458478123, 0.955957942875102, 0.871711503015831, 0.75764779211022, 0.43347541987896, 0.239236102672294, 0.0905989690218121, 0.992603611666709, 0.732277773553506, 0.124824774684384, 0.236234420444816, 0.145643977215514, 0.985394696006551, 0.348624687176198, 0.859788837376982, 0.513016011798754, 0.801681516924873, 0.275154813192785, 0.848803933011368, 0.195087121799588, 0.581892445450649, 0.920908353989944, 0.850194718455896, 0.314450426027179, 0.657697045942768, 0.876155759440735, 0.0912567693740129, 0.18735283985734, 0.18938291980885, 0.130232530413195, 0.627558624837548, 0.628286466933787, 0.387088863411918, 0.174361654324457, 0.140226917108521, 0.299076119903475, 0.999149184906855, 0.291054391534999, 0.655986906262115, 0.898435203358531, 0.598253333941102, 0.266524143982679, 0.196915044216439, 0.666080841096118, 0.94854798167944, 0.390371724264696, 0.0718863843940198, 0.834712795447558, 0.0605259400326759, 0.826661771861836, 0.565612607402727, 0.413377379998565, 0.600137619534507, 0.342170553514734, 0.909823346417397, 0.947757100686431, 0.8951659607701, 0.015578368678689, 0.965698821237311, 0.420144809409976, 0.770702671259642, 0.490726999938488, 0.949912334559485, 0.64058056473732, 0.188447463791817, 0.40679072146304, 0.384214348858222, 0.747413251083344, 0.822734112851322, 0.480520602781326, 0.638362947385758, 0.899067053338513, 0.680136315524578, 0.501831674482673, 0.0066291952971369, 0.88476404780522, 0.175478414166719, 0.252249646466225, 0.118403382599354, 0.960367433959618, 0.922156032640487, 0.341671304544434, 0.365229010116309, 0.0451114356983453, 0.70013706269674, 0.318060892401263, 0.145590789616108, 0.559851025929675, 0.739394953707233, 0.581532747019082, 0.355777028249577, 0.282780127367005, 0.858831661054865, 0.00168244424276054, 0.267336095450446, 0.989974816096947, 0.963782670674846, 0.627944659907371, 0.195020443759859, 0.407204198883846, 0.989680859725922, 0.24958622129634, 0.292091349838302, 0.43398087285459, 0.062773346202448, 0.445155154448003, 0.165399204939604, 0.17929062875919, 0.56858715438284, 0.394850595621392, 0.562004410894588, 0.734851227840409, 0.00192489195615053, 0.158529531443492, 0.571585651021451, 0.351021533831954, 0.959701438201591, 0.427358614513651, 0.492149695754051, 0.568287763977423, 0.109587070066482, 0.764116243924946, 0.0938205977436155, 0.448264261009172, 0.613908262457699, 0.0435343664139509, 0.88474809192121, 0.586762680672109, 0.960465908516198, 0.538429264444858, 0.723025594837964, 0.543292030924931, 0.0233540330082178, 0.892562168184668, 0.677063010400161, 0.574302285676822, 0.103276130510494, 0.108892483171076, 0.479781454894692, 0.55493300058879, 0.460705888690427, 0.208847751142457, 0.3936224039644, 0.535171133466065, 0.0251371883787215, 0.505127440439537, 0.750073783798143, 0.504015625920147, 0.182691366877407, 0.87111356947571, 0.595418989658356, 0.611793997231871, 0.17691142577678, 0.091652364237234, 0.336639854358509, 0.171722872881219, 0.911892907461151, 0.495836413931102, 0.478431262774393, 0.660523576429114, 0.232870045118034, 0.974255701061338, 0.48234897851944, 0.795795447425917, 0.732654300052673, 0.592981121968478, 0.318138463655487, 0.189951051725075, 0.928672429407015, 0.120756545802578, 0.779858061112463, 0.154845468932763, 0.396760487696156, 0.125549514079466, 0.726818086346611, 0.0541313621215522, 0.227597188204527, 0.146800782764331, 0.0417877961881459, 0.497385499998927, 0.515036009484902, 0.014642599504441, 0.423036300577223, 0.732843110570684, 0.94559886562638, 0.0606137453578413, 0.552549170097336, 0.235061477636918, 0.391912212129682, 0.718974858988076, 0.93672694824636, 0.689730311511084, 0.440235004527494, 0.0931316900532693, 0.925757465418428, 0.128739594016224, 0.43992504151538, 0.975139984162524, 0.344114036299288, 0.684841865673661, 0.152080860454589, 0.201821776572615, 0.107039859052747, 0.513220436871052, 0.481776180211455, 0.824480895418674, 0.606373729417101, 0.856656373944134, 0.607080389047042, 0.785386316478252, 0.688366798684001, 0.206433160463348, 0.299239170039073, 0.432681522332132, 0.0474248712416738, 0.431541851023212, 0.871345210587606, 0.333706641802564, 0.500017311424017, 0.252670951886103, 0.948419514112175, 0.344593436690047, 0.823379567591473, 0.43221532786265, 0.864578507607803, 0.509293694514781, 0.693103634519503, 0.742329803761095, 0.0756941975560039, 0.425651747733355, 0.456942624179646, 0.409554733196273, 0.795663218479604, 0.384906719904393, 0.92667450173758, 0.610716553637758, 0.404001674614847, 0.841238099848852, 0.475247261114419, 0.764195976080373, 0.578693123534322, 0.380617393180728, 0.00847509014420211, 0.966511242091656, 0.0436854159925133, 0.264942697482184, 0.333738765446469, 0.191795578459278, 0.0670015264768153, 0.338822203455493, 0.730378294596449, 0.538147605257109, 0.223518826533109, 0.726159904384986, 0.49582930910401, 0.758527630707249, 0.295817972626537, 0.875250783050433, 0.461311652790755, 0.513591029215604, 0.757304381579161, 0.264806606341153, 0.137404415290803, 0.894817249150947, 0.0923038998153061, 0.256727543193847, 0.480237694224343, 0.763808652060106, 0.0753335312474519, 0.59352487558499, 0.499561764067039, 0.448204163927585, 0.573697270127013, 0.639490750618279, 0.503303921315819, 0.000977888004854321, 0.778285765787587, 0.508507502265275, 0.441587988985702, 0.21124866255559, 0.229117087787017, 0.56431383965537, 0.17917018616572, 0.349408986745402, 0.975045442115515, 0.28076504310593, 0.498159982031211, 0.426194148836657, 0.726045297225937, 0.00418231869116426, 0.682121147401631, 0.673467074753717, 0.0523294759914279, 0.79334574053064, 0.375597400357947, 0.00557474559172988, 0.250218925531954, 0.355455122189596, 0.413540898356587, 0.14917966700159, 0.0392227005213499, 0.421770803630352, 0.110657557612285, 0.514830087777227, 0.153380839619786, 0.114125763997436, 0.847503160592169, 0.0700877513736486, 0.458432670217007, 0.124983120476827, 0.0103925531730056, 0.912550368579105, 0.55539080966264, 0.603329277830198, 0.0506194119807333, 0.0631316965445876, 0.568029290065169, 0.330412819050252, 0.553934115916491, 0.300009503727779, 0.987608019961044, 0.208860547980294, 0.0494392786640674, 0.538735619047657, 0.49211192666553, 0.293202015804127, 0.0520175187848508, 0.857808666070923, 0.694678931962699, 0.779880541842431, 0.827792447991669, 0.598644117591903, 0.754933068528771, 0.620421862462536, 0.172728141769767, 0.219826468266547, 0.529761329526082, 0.777253289474174, 0.777492270572111, 0.728147012414411, 0.731365331215784, 0.0567409845534712, 0.24866332905367, 0.563594859093428, 0.407914256444201, 0.984037824207917, 0.21676425402984, 0.707525141537189, 0.903782803099602, 0.11379867605865, 0.1112115168944, 0.946625158190727, 0.1983475855086, 0.989582038717344, 0.451426580548286, 0.12849257956259, 0.592543999664485, 0.0225500364322215, 0.667268652934581, 0.777086738264188, 0.24962060013786, 0.00923931645229459, 0.619613227434456, 0.634680049493909, 0.429967417614534, 0.398219477152452, 0.231074950657785, 0.879571032011881, 0.105807145358995, 0.174104934092611, 0.379845224088058, 0.212897182675079, 0.474310318706557, 0.919971000170335, 0.517570148222148, 0.245078759500757, 0.287281929515302, 0.228272090898827, 0.359817468794063, 0.309022984467447, 0.416876767529175, 0.861804463900626, 0.562545794062316, 0.937657236354426, 0.182472004089504, 0.710288329515606, 0.729541229316965, 0.807484164368361, 0.450249377638102, 0.244834008393809, 0.333238667575642, 0.759248808724806, 0.450393822276965, 0.150968797039241, 0.618215106194839, 0.639010927872732, 0.75919346138835, 0.196810914902017, 0.754101704107597, 0.708136610453948, 0.423628628952429, 0.614805556600913, 0.7808218204882, 0.201938150217757, 0.537899172632024, 0.108181342016906, 0.521604582900181, 0.195530225289986, 0.255944428034127, 0.815744154620916, 0.43829193036072, 0.890118289273232, 0.00882343971170485, 0.63514897832647, 0.388355851406232, 0.532619923353195, 0.0197334485128522, 0.702496020821854, 0.585788806667551, 0.0881375211756676, 0.277592495083809, 0.368738688062876, 0.173050748882815, 0.424155664397404, 0.963274590438232, 0.889423114713281, 0.656518345931545, 0.707562661729753, 0.257933119544759, 0.0956805478781462, 0.49477723822929, 0.0159611753188074, 0.272638842230663, 0.997114951955155, 0.618016375461593, 0.00357068330049515, 0.667160366661847, 0.370999595848843, 0.571668114280328, 0.0291409529745579, 0.80301069188863, 0.0945478989742696, 0.0940861606504768, 0.675001717405394, 0.776591819012538, 0.813395036151633, 0.0858693576883525, 0.514063544105738, 0.872468322049826, 0.180116092786193, 0.109923663083464, 0.263827688060701, 0.774187913164496, 0.783360851928592, 0.11040881392546, 0.291019576834515, 0.871981124626473, 0.689173844177276, 0.748424412915483, 0.889865266624838, 0.88271840242669, 0.130977960769087, 0.304652629187331, 0.741835539927706, 0.838753486983478, 0.333458179840818, 0.605420218082145, 0.562042777426541, 0.890568768838421, 0.242040487239137, 0.321981147862971, 0.254970064852387, 0.32877646619454, 0.872093298938125, 0.923149587586522, 0.246006071101874, 0.806714233709499, 0.297129615442827, 0.644172792090103, 0.0607197468634695, 0.143364455085248, 0.640495799481869, 0.324178807204589, 0.173055627150461, 0.657726444769651, 0.998402117984369, 0.5536406962201, 0.76367201982066, 0.617371889296919, 0.437294452916831, 0.803111511049792, 0.387465678155422, 0.867239263607189, 0.00506599759683013, 0.663985420484096, 0.493535242509097, 0.648932306328788, 0.201122348662466, 0.836408515227959, 0.415261517977342, 0.404092440847307, 0.222488953964785, 0.826340075116605, 0.0710694866720587, 0.548927166732028, 0.348169501172379, 0.202727064490318, 0.880017616786063, 0.286154227796942, 0.649371471023187, 0.628308420535177, 0.211420459439978, 0.223315715091303, 0.505020718555897, 0.425270397448912, 0.962132125161588, 0.405493619851768, 0.77729283249937, 0.366674898425117, 0.617836914490908, 0.3305455413647, 0.457621919224039, 0.744901102501899, 0.00507403910160065, 0.182897999649867, 0.767346408683807, 0.141684390371665, 0.707672859542072, 0.0922362632118165, 0.4933511831332, 0.842031640233472, 0.58036180655472, 0.228449511341751, 0.963139311643317, 0.0242189322598279, 0.0392338377423584, 0.313830916071311, 0.985349732916802, 0.587671924382448, 0.794778418261558, 0.579397822031751, 0.516793441725895, 0.492263886611909, 0.580878335284069, 0.516345593612641, 0.444950931472704, 0.830271284561604, 0.0851976308040321, 0.993077589664608, 0.0143972944933921, 0.322192730847746, 0.169855875894427, 0.0271965477149934, 0.304600374074653, 0.309511209838092, 0.755403019255027, 0.874933535698801, 0.833916679490358, 0.0236036500427872, 0.0531821786426008, 0.19065536535345, 0.727464212570339, 0.120582580333576, 0.80122219119221, 0.242939934367314, 0.758517019217834, 0.310826347209513, 0.593860533321276, 0.528792369179428, 0.270802656188607, 0.224344751564786, 0.634950479259714, 0.832030734512955, 0.472376765683293, 0.33105677482672, 0.988094160798937, 0.167539009358734, 0.23870588093996, 0.982387918280438, 0.828951637027785, 0.389749128138646, 0.305287709226832, 0.432742834789678, 0.144650992704555, 0.614642545348033, 0.451169074513018, 0.658910225378349, 0.430983036523685, 0.83020127331838, 0.705027302494273, 0.888733638683334, 0.8882754992228, 0.22748668375425, 0.146409249166027, 0.0526201438624412, 0.852903706021607, 0.561012061079964, 0.159478186629713, 0.101119097787887, 0.539874112699181, 0.429218955338001, 0.678184928838164, 0.768589794170111, 0.847207189071923, 0.943748542340472, 0.257880531251431, 0.00808262638747692, 0.535703629488125, 0.865154469152912, 0.739678252255544, 0.995307428762317, 0.90580068086274, 0.151351331500337, 0.684818569337949, 0.0182705982588232, 0.294393178308383, 0.442572126863524, 0.164868998108432, 0.974081420106813, 0.00341265392489731, 0.0194605847354978, 0.451566770672798, 0.425534520531073, 0.636833301046863, 0.943338027456775, 0.240178662119433, 0.869260625448078, 0.0893007437698543, 0.836526485858485, 0.366515340283513, 0.76990664168261, 0.361778888385743, 0.735776759684086, 0.260605301940814, 0.283374883234501, 0.38761712773703, 0.117115026572719, 0.680337051395327, 0.176960123702884, 0.365721338428557, 0.0646280082874, 0.713808514643461, 0.603827798971906, 0.4535203485284, 0.259609014727175, 0.122703684959561, 0.319839491741732, 0.451315156649798, 0.944175576092675, 0.132871716748923, 0.348764314083382, 0.620079026324674, 0.388564751017839, 0.927830272587016, 0.36705118813552, 0.927299031522125, 0.0656410346273333, 0.324557224521413, 0.232919392408803, 0.892451357329264, 0.980947826290503, 0.752221622969955, 0.226514676120132, 0.497544193640351, 0.346686812583357, 0.305884210858494, 0.602004529908299, 0.559735298389569, 0.160536314826459, 0.294419017387554, 0.304863810306415, 0.27750161383301, 0.445604252861813, 0.523564582457766, 0.596700027119368, 0.102376724826172, 0.700025144033134, 0.628599957097322, 0.380060403840616, 0.727101230761036, 0.914448974421248, 0.860656959237531, 0.888843514490873, 0.246038268553093, 0.824374446179718, 0.273987544234842, 0.601445969427004, 0.983815859770402, 0.949815422296524, 0.194700192660093, 0.431881197495386, 0.685920001938939, 0.678186031291261, 0.501365888630971, 0.407502243062481, 0.340395932318643, 0.170964313903823, 0.584699821425602, 0.923015719745308, 0.171223700046539, 0.749666122952476, 0.912877196678892, 0.515016868012026, 0.410148344701156, 0.466544013237581, 0.997078932821751, 0.5124269539956, 0.461433466523886, 0.971717710606754, 0.61252259509638, 0.637274717912078, 0.488449794240296, 0.198643951909617, 0.807106101885438, 0.208871228154749, 0.888901603640988, 0.978113847551867, 0.476528087863699, 0.131352873751894, 0.744060856755823, 0.803891077404842, 0.243858740665019, 0.175795264542103, 0.832298686960712, 0.622976816026494, 0.361992773367092, 0.924183999653906, 0.21075353003107, 0.409884131280705, 0.9755570422858, 0.568534805905074, 0.408254196168855, 0.841691243695095, 0.976046474417672, 0.390583539614454, 0.358005528803915, 0.0740889976732433, 0.50462551927194, 0.49653752357699, 0.94238522532396, 0.87632129364647, 0.825918801594526, 0.672011869493872, 0.182931032497436, 0.396721585188061, 0.342608441598713, 0.454972980543971, 0.194042477989569, 0.514042734634131, 0.423460208810866, 0.445125637343153, 0.109152854187414, 0.591349379392341, 0.389637323096395, 0.914978586602956, 0.616037359228358, 0.118341326480731, 0.720446364954114, 0.244132907828316, 0.904649051139131, 0.502542862202972, 0.230574611807242, 0.0918201997410506, 0.234785120701417, 0.649235995486379, 0.459199648816139, 0.378687093267217, 0.563709776615724, 0.431355747859925, 0.0149041272234172, 0.722359551116824, 0.203736533410847, 0.843450399348512, 0.472629844909534, 0.946098575135693, 0.966251600533724, 0.423432383919135, 0.986894264351577, 0.274473239900544, 0.76073359255679, 0.963032126892358, 0.240419543813914, 0.478026159340516, 0.17293406650424, 0.90620302525349, 0.102691826876253, 0.657071364345029, 0.869831036543474, 0.870461737504229, 0.128787311958149, 0.857868138700724, 0.195634681265801, 0.876948745222762, 0.931046638637781, 0.81831272947602, 0.921146506210789, 0.518934618448839, 0.592332907021046, 0.0795916824135929, 0.7475341190584, 0.261882691411301, 0.148158527910709, 0.128740918822587, 0.00741095212288201, 0.347073504934087, 0.336041504517198, 0.969523224281147, 0.377001367509365, 0.930598756764084, 0.315178429475054, 0.875497439876199, 0.299269089009613, 0.745981895597652, 0.633585668867454, 0.270798329263926, 0.621838200604543, 0.480157135985792, 0.180805857758969, 0.349608856718987, 0.543966637691483, 0.0429890977684408, 0.747907448327169, 0.493498139781877, 0.0134759899228811, 0.625784709351137, 0.870004229946062, 0.178677874850109, 0.791400587186217, 0.614779702853411, 0.578909436240792, 0.229718888644129, 0.604502826696262, 0.872612604871392, 0.533096185186878, 0.744284072192386, 0.48945870436728, 0.301067287800834, 0.78910418623127, 0.840742808068171, 0.224437016062438, 0.881118217017502, 0.299345909850672, 0.203603680245578, 0.906081310007721, 0.49244039109908, 0.527860375121236, 0.040566970128566, 0.68186144484207, 0.710349631728604, 0.227597292279825, 0.488833735464141, 0.0145808586385101, 0.672536678845063, 0.765597344143316, 0.80598333152011, 0.842263863421977, 0.161038025515154, 0.89322661748156, 0.233131722779945, 0.0143840829841793, 0.000754152657464147, 0.412308746716008, 0.845289222430438, 0.247376619605348, 0.810814795549959, 0.0916323547717184, 0.335703815100715, 0.636298694647849, 0.161259726155549, 0.761784509755671, 0.883874537423253, 0.642262677894905, 0.179874126100913, 0.29484636709094, 0.307732861954719, 0.0828014169819653, 0.355123441666365, 0.881528760073707), actual = c(0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.754962249426171, 0.132685066200793, 0.0214571761898696, 0.0311931276228279, 0.37108744843863, 0.6087897492107, 0.873564527835697, 0.341455447487533, 0.864530329592526, 0.281694698845968, 0.406810106709599, 0.660866797203198, 0.200803185114637, 0.336926651652902, 0.355347598204389, 0.156645421404392, 0.562487736111507, 0.9373552815523, 0.982533112633973, 0.534236210631207, 0.911504037678242, 0.758615577593446, 0.110702075529844, 0.808069208404049, 0.379296046681702, 0.722280216636136, 0.76175744831562, 0.348066494101658, 0.23428487055935, 0.74404037813656, 0.183191269403324, 0.223210882395506, 0.0754236672073603, 0.592405981384218, 0.601156858727336, 0.0361738812644035, 0.217197973513976, 0.130867335014045, 0.117819632403553, 0.82679190300405, 0.60291959065944, 0.864306289004162, 0.132226389367133, 0.854048894718289, 0.951326867099851, 0.452719093766063, 0.676642122911289, 0.929277641931549, 0.201363323023543, 0.648554368875921, 0.412200199905783, 0.572590808849782, 0.196798804216087, 0.231667596148327, 0.947052046190947, 0.277816675137728, 0.754651495721191, 0.262576997745782, 0.837298149708658, 0.61937146121636, 0.891436998266727, 0.468928356189281, 0.638735354412347, 0.101613801671192, 0.373230731114745, 0.663178937975317, 0.84864161349833, 0.0806232099421322, 0.536123180994764, 0.0179828964173794, 0.977655903669074, 0.455069354502484, 0.70226266165264, 0.193864058703184, 0.540234672138467, 0.822998405434191, 0.0879073985852301, 0.0270173873286694, 0.671470671426505, 0.855909805977717, 0.927265778649598, 0.586976064834744, 0.997798324329779, 0.562158285407349, 0.0541052748449147, 0.790475640445948, 0.419829684775323, 0.885494893882424, 0.717261974699795, 0.502215886022896, 0.0357010124716908, 0.813421878498048, 0.592392867198214, 0.763182389782742, 0.889828151324764, 0.858651013346389, 0.204159409739077, 0.0415465305559337, 0.555067068431526, 0.458018211647868, 0.414596838643774, 0.816047390690073, 0.294556581182405, 0.421841554576531, 0.998315325705335, 0.849310165503994, 0.019374976400286, 0.0305166377220303, 0.171333615202457, 0.799628352047876, 0.416360178729519, 0.165041126776487, 0.840910330647603, 0.399574243696406, 0.453132037771866, 0.0596601236611605, 0.397423067130148, 0.817589501617476, 0.6944410353899, 0.642686699517071, 0.0787031392101198, 0.489925056230277, 0.208719278685749, 0.684071848168969, 0.246734193526208, 0.843741197604686, 0.923426605993882, 0.630496578523889, 0.845708502922207, 0.281709410715848, 0.402258559595793, 0.68597480840981, 0.967795077944174, 0.66219778987579, 0.32067319448106, 0.484644707525149, 0.416609379928559, 0.126568890875205, 0.175416003679857, 0.319959190208465, 0.635542009724304, 0.885833989828825, 0.862006564391777, 0.237598890438676, 0.674721467541531, 0.766074177576229, 0.895803815685213, 0.527197161223739, 0.573490129318088, 0.992163812043145, 0.338035006076097, 0.00107767479494214, 0.861195669742301, 0.99823805061169, 0.132823590189219, 0.413978235330433, 0.686710680834949, 0.260584762319922, 0.787106965435669, 0.0807143449783325, 0.94640559121035, 0.618190953973681, 0.406772780697793, 0.263069764012471, 0.461943653179333, 0.0899391216225922, 0.204920787829906, 0.0826278850436211, 0.420106585836038, 0.0207044894341379, 0.445098555414006, 0.399643331067637, 0.401188868563622, 0.779219570336863, 0.796101979911327, 0.242056191666052, 0.176316885044798, 0.644789448240772, 0.947168268961832, 0.780890108086169, 0.309441871009767, 0.932981422170997, 0.381271421443671, 0.551071044523269, 0.211657229810953, 0.839983659563586, 0.520944925723597, 0.627602343447506, 0.376259338110685, 0.395456904312596, 0.660570190520957, 0.848912462592125, 0.413498609093949, 0.641091976547614, 0.708299761405215, 0.466558170970529, 0.765544437570497, 0.706805642461404, 0.691481506917626, 0.293228038121015, 0.207067188341171, 0.249698005383834, 0.940138085512444, 0.29468309902586, 0.976801350712776, 0.804110781522468, 0.695043537067249, 0.783054896863177, 0.724812642438337, 0.375795488012955, 0.437980354530737, 0.919426952954382, 0.344140102155507, 0.721373289590701, 0.86133532342501, 0.997512152651325, 0.58625375945121, 0.839988996973261, 0.933821993181482, 0.95061707100831, 0.314740929054096, 0.521034938981757, 0.014084339607507, 0.118985927430913, 0.815023053670302, 0.211415915051475, 0.368022182956338, 0.681176419835538, 0.367485424038023, 0.468315284699202, 0.739310655277222, 0.517207938013598, 0.960209380835295, 0.411323901731521, 0.20360124623403, 0.0259302840568125, 0.474839743226767, 0.0116984294727445, 0.734752660617232, 0.240566151682287, 0.292693223804235, 0.312738243956119, 0.515484969364479, 0.120809535728768, 0.565610871417448, 0.898485622834414, 0.893438902217895, 0.839242467889562, 0.717722930014133, 0.352856595302001, 0.347004774259403, 0.791036127600819, 0.133946795482188, 0.0686658581253141, 0.346860635327175, 0.651698319707066, 0.442245902726427, 0.604501663707197, 0.42069887672551, 0.20824971748516, 0.251475076889619, 0.457427399232984, 0.252967544831336, 0.443571016192436, 0.00233758799731731, 0.190317516913638, 0.426044536754489, 0.48293644329533, 0.934835736639798, 0.218599590007216, 0.325623096432537, 0.130601252196357, 0.530534722842276, 0.95615404099226, 0.356284614419565, 0.345872638048604, 0.40915888058953, 0.852933375863358, 0.0446312630083412, 0.706984600517899, 0.414325835648924, 0.216973330592737, 0.534994187299162, 0.198739880463108, 0.0139784412458539, 0.724894990213215, 0.539284620899707, 0.891309102997184, 0.822146094404161, 0.544926994480193, 0.837274253368378, 0.433921710588038, 0.0398945510387421, 0.00725216045975685, 0.162786038359627, 0.57048646430485, 0.343309990828857, 0.459390287986025, 0.880073568085209, 0.294620576081797, 0.769311589887366, 0.186188709223643, 0.301876226905733, 0.40327766048722, 0.397118548862636, 0.792037378530949, 0.582340413937345, 0.736929765436798, 0.779017042601481, 0.0809617817867547, 0.858656670898199, 0.272883453639224, 0.236575785791501, 0.0890757406596094, 0.610618126112968, 0.554921263828874, 0.800966496113688, 0.524243202526122, 0.808449850650504, 0.669351305579767, 0.823347354540601, 0.680034450255334, 0.679285305552185, 0.211000747280195, 0.475458631059155, 0.836235595634207, 0.303788753692061, 0.0974186731036752, 0.369493928737938, 0.135217234725133, 0.30151364277117, 0.971895388094708, 0.631912175333127, 0.47092240373604, 0.467068093828857, 0.184201356256381, 0.755781409563497, 0.0608622538857162, 0.867034818278626, 0.762950614560395, 0.194925687042996, 0.102918421151116, 0.975166266551241, 0.541952325031161, 0.428354703355581, 0.297591837821528, 0.360334273660555, 0.0737996068783104, 0.458616176852956, 0.380867247469723, 0.210745889693499, 0.588451374787837, 0.799468436976895, 0.1302116853185, 0.457708096830174, 0.37411332433112, 0.55482278438285, 0.522480322280899, 0.257192247081548, 0.915535778971389, 0.0413645615335554, 0.68819476547651, 0.309856554260477, 0.245010945945978, 0.389962851069868, 0.880415399093181, 0.394415379501879, 0.554443422239274, 0.608121007680893, 0.183655625209212, 0.119744806084782, 0.798768220702186, 0.140265758847818, 0.481148412916809, 0.2639424004592, 0.676550675882027, 0.89126879023388, 0.279458449454978, 0.99821522529237, 0.495093673933297, 0.393188353162259, 0.644902628380805, 0.223438819637522, 0.257310273591429, 0.0988306212238967, 0.961879282491282, 0.591950163478032, 0.727934378897771, 0.30350111075677, 0.472294552717358, 0.755782925290987, 0.270625886507332, 0.41920671518892, 0.586749933194369, 0.22640467225574, 0.257757667219266, 0.880703550763428, 0.912534300005063, 0.572221976937726, 0.489657329861075, 0.711260867770761, 0.290741693926975, 0.875601920997724, 0.0205898662097752, 0.782044317340478, 0.816758941859007, 0.485877371858805, 0.833557714242488, 0.804361337097362, 0.823168118018657, 0.323853453155607, 0.962907359004021, 0.510178663069382, 0.0443144508171827, 0.796154369134456, 0.74516949057579, 0.789826899999753, 0.425149316899478, 0.501493294490501, 0.839258925523609, 0.284971047891304, 0.34903074009344, 0.356835737125948, 0.482654102845117, 0.310062436386943, 0.821647999342531, 0.867418338777497, 0.750065328320488, 0.401930406456813, 0.335573024582118, 0.833379876567051, 0.18684041290544, 0.360822014510632, 0.283695445396006, 0.0468338334467262, 0.145307250786573, 0.393436731770635, 0.00283416314050555, 0.336895487969741, 0.0124647377524525, 0.183011097367853, 0.189903337275609, 0.820932957110927, 0.660787503002211, 0.534866307862103, 0.261576964752749, 0.833485843380913, 0.953830800717697, 0.810203492641449, 0.703040716936812, 0.182745023397729, 0.496705999132246, 0.0847508597653359, 0.169720595004037, 0.952114598825574, 0.0644094846211374, 0.187176185194403, 0.859008980216458, 0.542653847951442, 0.524917643284425, 0.132794436998665, 0.841731026535854, 0.460362837882712, 0.618603854905814, 0.0803159670904279, 0.33063528756611, 0.483616817509755, 0.500358528224751, 0.139481038087979, 0.118751945206895, 0.849000166170299, 0.379106594249606, 0.174959697527811, 0.507162440102547, 0.15595520241186, 0.0643125895876437, 0.0792997491080314, 0.555262855719775, 0.0970915358047932, 0.219785890076309, 0.762458481127396, 0.739774220390245, 0.804275318747386, 0.154555402696133, 0.124188252724707, 0.799627799773589, 0.521943091182038, 0.316208606120199, 0.233371592359617, 0.593053756281734, 0.221609525149688, 0.462189205922186, 0.073846586747095, 0.413742355071008, 0.482608171878383, 0.693696758709848, 0.156406844034791, 0.923639398766682, 0.404950670897961, 0.0855093956924975, 0.685520310653374, 0.678066204534844, 0.535323498072103, 0.123473756248131, 0.405812355456874, 0.963988753501326, 0.865197021514177, 0.989571653772146, 0.291450886754319, 0.784245166461915, 0.619239983381703, 0.0914580242242664, 0.197074436815456, 0.0991156687960029, 0.974926596041769, 0.00528142228722572, 0.946832921355963, 0.794790945248678, 0.424180187517777, 0.425383554771543, 0.823408173397183, 0.741251219063997, 0.920135132735595, 0.731587630230933, 0.475290835369378, 0.624358572997153, 0.664764505811036, 0.0655632379930466, 0.601710018934682, 0.316776387626305, 0.910055433632806, 0.0464452472515404, 0.452059453353286, 0.434997930889949, 0.104075973387808, 0.839991080807522, 0.574121955549344, 0.464823849964887, 0.0763848994392902, 0.968895534984767, 0.236444991547614, 0.910935632418841, 0.441378109157085, 0.844549528323114, 0.837453079177067, 0.306358754169196, 0.372765195090324, 0.278117963345721, 0.734935146989301, 0.65036468510516, 0.219651677878574, 0.960304379463196, 0.937941479030997, 0.687057570321485, 0.767151058418676, 0.265132369473577, 0.0759228894021362, 0.555635241558775, 0.71604901435785, 0.275231721345335, 0.280902422266081, 0.212997151073068, 0.956354178953916, 0.823789662914351, 0.950799788115546, 0.0659945432562381, 0.2933260188438, 0.7762974309735, 0.783336439868435, 0.458911170018837, 0.596193538745865, 0.952319821808487, 0.689989881357178, 0.0611461461521685, 0.586165876826271, 0.683481096290052, 0.337322106352076, 0.710446609416977, 0.154328410513699, 0.94472940871492, 0.306524229468778, 0.648713312344626, 0.0721474750898778, 0.724112971220165, 0.14837013464421, 0.831568174064159, 0.282319550868124, 0.396638519596308, 0.890910838264972, 0.174657220253721, 0.859681763220578, 0.0432850080542266, 0.77993385354057, 0.841268370626494, 0.145291790366173, 0.911118480376899, 0.311985693173483, 0.676699575502425, 0.836418931372464, 0.977840786566958, 0.413668513298035, 0.0796897469554096, 0.584752671187744, 0.491445231018588, 0.158104622038081, 0.842387928860262, 0.888755855383351, 0.280352438334376, 0.0425687765236944, 0.0722129282075912, 0.201807521050796, 0.929171522613615, 0.407017460092902, 0.0207342037465423, 0.0137804683763534, 0.624601892894134, 0.428407954983413, 0.388107931707054, 0.152528431965038, 0.939300283323973, 0.657127618324012, 0.713982358342037, 0.787979938089848, 0.0456434467341751, 0.0869422433897853, 0.437283435603604, 0.766174303600565, 0.702980798203498, 0.328449685126543, 0.254320837091655, 0.232890106504783, 0.529887686250731, 0.858251673402265, 0.891638060566038, 0.853619711007923, 0.485483623109758, 0.43692403123714, 0.316335140028968, 0.495007500983775, 0.925124336499721, 0.398114950628951, 0.234059053706005, 0.379389526555315, 0.741351113654673, 0.440635880455375, 0.205532220890746, 0.928000486455858, 0.994942751247436, 0.0827509043738246, 0.270670740166679, 0.624442738713697, 0.370574478292838, 0.0187001540325582, 0.274001910351217, 0.609464508015662, 0.628908675629646, 0.657028979388997, 0.887561597162858, 0.374038046225905, 0.271224809577689, 0.1346954414621, 0.271649493137375, 0.649006454041228, 0.233826696872711, 0.500379588687792, 0.942730586510152, 0.690699894214049, 0.989228627644479, 0.36743511701934, 0.126515429699793, 0.185395603301004, 0.437740752007812, 0.0620097443461418, 0.438721900805831, 0.000589912757277489, 0.283154103672132, 0.44967214949429, 0.539968436118215, 0.0611403696238995, 0.759065968450159, 0.972009486053139, 0.252417966257781, 0.0517126459162682, 0.369742979761213, 0.362735068658367, 0.375995533773676, 0.846414015861228, 0.903908740496263, 0.364329840289429, 0.98593678092584, 0.150161722907797, 0.204421801492572, 0.315115064848214, 0.310436042025685, 0.54494786914438, 0.683466287096962, 0.313104300759733, 0.768034709850326, 0.17179629765451, 0.476326169446111, 0.537133234087378, 0.227867004228756, 0.252165507525206, 0.957884557312354, 0.093629238428548, 0.95466060214676, 0.234657989582047, 0.843165446305647, 0.736869102576748, 0.24904672591947, 0.247364209964871, 0.53884691442363, 0.473282627295703, 0.459163626190275, 0.258803543401882, 0.896768743405119, 0.818418883020058, 0.748121270211414, 0.320617051096633, 0.743334874976426, 0.787065841723233, 0.652753064641729, 0.0488581517711282, 0.358596400823444, 0.0726789063774049, 0.149300223216414, 0.363278931705281, 0.596705817384645, 0.448919540969655, 0.156575797358528, 0.174629051005468, 0.617407873738557, 0.430267909308895, 0.586960633983836, 0.457926504313946, 0.563388651702553, 0.0918506786692888, 0.874157186364755, 0.745841363910586, 0.0291785176377743, 0.165784746874124, 0.23153902287595, 0.799632784212008, 0.80045567988418, 0.968691506423056, 0.196303070988506, 0.523661050014198, 0.215383989969268, 0.396582623710856, 0.373583944514394, 0.600238279206678, 0.207723492290825, 0.210544881178066, 0.310006387764588, 0.975586319109425, 0.278426877921447, 0.240950037026778, 0.0495633308310062, 0.780981200514361, 0.2718488893006, 0.965636740904301, 0.255128050222993, 0.475706724915653, 0.465936198597774, 0.536126364488155, 0.0303886148612946, 0.00904769380576909, 0.976971724070609, 0.52774966834113, 0.175015406683087, 0.240281759761274, 0.37617775821127, 0.718493678141385, 0.159584995126352, 0.839446405181661, 0.518482028972358, 0.523120517609641, 0.0612899227999151, 0.953741739038378, 0.494232170982286, 0.299054854083806, 0.3215735796839, 0.886399049311876, 0.687575902557001, 0.595135164214298, 0.534132891334593, 0.377301646629348, 0.295220714528114, 0.824122667312622, 0.386810271302238, 0.46602792548947, 0.502571142744273, 0.0275159326847643, 0.855979087995365, 0.394415259361267, 0.66880876943469, 0.286518510663882, 0.703288921155035, 0.909905741922557, 0.482302644988522, 0.618072894169018, 0.937729975441471, 0.312440904555842, 0.665120480582118, 0.891072019701824, 0.7575912931934, 0.189815272111446, 0.80279399594292, 0.403007111977786, 0.112475694157183, 0.401964175282046, 0.734072093386203, 0.011599926976487, 0.194310583174229, 0.716895660152659, 0.630588712869212, 0.984184815082699, 0.27319520409219, 0.420970570063218, 0.401669973973185, 0.626153575023636, 0.468379428610206, 0.0623241544235498, 0.535719451261684, 0.00938250101171434, 0.0323439412750304, 0.772903016302735, 0.838810488115996, 0.251342195086181, 0.736817829310894, 0.712556465994567, 0.328902672510594, 0.829779686639085, 0.125293884892017, 0.852771409088746, 0.266158617567271, 0.0316743773873895, 0.782306514447555, 0.70720084849745, 0.41025125537999, 0.993782072560862, 0.717811059439555, 0.657738904235885, 0.491708368528634, 0.152852487517521, 0.747963017784059, 0.59845435176976, 0.130087576573715, 0.248942216392606, 0.0919888145290315, 0.312387069920078, 0.980522795580328, 0.682165225734934, 0.347577383974567, 0.660695116501302, 0.947079666424543, 0.592637636931613, 0.808706965530291, 0.905255305813625, 0.713226166553795, 0.785025741439313, 0.0250550645869225, 0.189160379348323, 0.00944804237224162, 0.59489083965309, 0.501591209555045, 0.632648948812857, 0.594013348920271, 0.32845940371044, 0.548319486668333, 0.245759726269171, 0.775033091660589, 0.613817256642506, 0.777066848939285, 0.924323311541229, 0.387695215875283, 0.52122691180557, 0.0523970811627805, 0.261659845244139, 0.673075395636261, 0.11486284667626, 0.42924763308838, 0.347878718515858, 0.387759200530127, 0.864613679936156, 0.313331987475976, 0.472535834414884, 0.662030744133517, 0.0777701728511602, 0.953692942857742, 0.0615598692093045, 0.0703973097261041, 0.644247290212661, 0.950591987697408, 0.847898351028562, 0.285136505961418, 0.244237698614597, 0.725537044461817, 0.371868761954829, 0.433893900597468, 0.82413102989085, 0.86562753864564, 0.901342392899096, 0.0174278516788036, 0.124666254268959, 0.932023116387427, 0.899888148996979, 0.539852513698861, 0.936668258858845, 0.767765648430213, 0.972031750483438, 0.12759614107199, 0.87341947783716, 0.738317348994315, 0.991644144291058, 0.525290976511315, 0.917075187899172, 0.00805971352383494, 0.825418079039082, 0.844087673118338, 0.631323665380478, 0.899504894390702, 0.189054476795718, 0.117631731322035, 0.340717596001923, 0.301188428886235, 0.962240171385929, 0.350029700901359, 0.0350526922848076, 0.869892306625843, 0.150640804087743, 0.00590812205336988, 0.419224388198927, 0.720309484982863, 0.935179687570781, 0.582711809780449, 0.276567449094728, 0.328746677609161, 0.151025800267234, 0.287338309688494, 0.0521411644294858, 0.543096114881337, 0.387272034073249, 0.642282930668443, 0.979066597297788, 0.555690587032586, 0.993989021284506, 0.950300248572603, 0.661176170920953, 0.152426039800048, 0.884294432820752, 0.423775085946545, 0.469105489552021, 0.241852237377316, 0.0685343174263835, 0.698924507247284, 0.396203014533967, 0.701154673704877, 0.393378956476226, 0.0307144832331687, 0.977974398992956, 0.476410410134122, 0.157637000083923, 0.584274036809802, 0.926045883679762, 0.493777793832123, 0.137997763697058, 0.164449085714296, 0.968185652978718, 0.677444187924266, 0.241831120569259, 0.456109498394653, 0.949956724653021, 0.314150095451623, 0.69932885770686, 0.143734187353402, 0.597156836884096, 0.564528892049566, 0.424847566289827, 0.817563744494691, 0.534121020929888, 0.936842138180509, 0.809629861265421, 0.981073004659265, 0.62309005134739, 0.171884892508388, 0.599706296343356, 0.0737552712671459, 0.324512123363093, 0.595465854508802, 0.622749016853049, 0.281722405459732, 0.181302106939256, 0.29423041571863, 0.239586413837969, 0.888523889239877, 0.909916238393635, 0.901355616049841, 0.817037930479273, 0.741819407092407), actual = c(0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"