Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>        0.0000000        1.0000000        0.4775943        0.4775943 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>        0.5000000        0.4775943        0.4775943        0.5100487 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.653841512976214, 0.0864882674068213, 0.429085751296952, 0.633056990569457, 0.713701906381175, 0.260900755645707, 0.232571922242641, 0.773499462287873, 0.50952209928073, 0.419372700387612, 0.936496928799897, 0.560205965535715, 0.732280182186514, 0.571301782270893, 0.352733991574496, 0.731878438498825, 0.612969329813495, 0.321308071259409, 0.760427175322548, 0.479639552999288, 0.988754460122436, 0.363412248669192, 0.592175291152671, 0.489546798635274, 0.477947404840961, 0.79654958541505, 0.733001176966354, 0.29620934301056, 0.828472294844687, 0.0159411972854286, 0.619909179862589, 0.102612493559718, 0.93957769847475, 0.254862203728408, 0.504759664647281, 0.424687992082909, 0.315262225223705, 0.700105306925252, 0.147020950214937, 0.484531702008098, 0.692711194278672, 0.974586619064212, 0.474046175833791, 0.0550125979352742, 0.833576294127852, 0.654759855242446, 0.722057026810944, 0.543311273213476, 0.236450431169942, 0.391411633230746, 0.0266984177287668, 0.547442573355511, 0.502792726270854, 0.556801639031619, 0.00974425696767867, 0.178696143208072, 0.796668467344716, 0.676750939106569, 0.54724909667857, 0.482270973268896, 0.896655945573002, 0.239702137187123, 0.0150550818070769, 0.521017141174525, 0.887501505436376, 0.458095179172233, 0.7713797907345, 0.175465799402446, 0.349278117762879, 0.544055553618819, 0.109122283291072, 0.287442483240739, 0.186640501488, 0.827489324379712, 0.15171403484419, 0.598972936160862, 0.234528059372678, 0.194213105831295, 0.580422236351296, 0.279153428273275, 0.353330867132172, 0.926364945946261, 0.974249333143234, 0.784060092875734, 0.291859378106892, 0.74154804344289, 0.740668127080426, 0.399928901810199, 0.320311416406184, 0.952472251607105, 0.262338389875367, 0.0426247571595013, 0.93732066033408, 0.604354817187414, 0.530173144303262, 0.559022812172771, 0.922319148899987, 0.00534512661397457, 0.329258188838139, 0.591234991792589, 0.0278647753875703, 0.145377360284328, 0.824785843025893, 0.671325333882123, 0.756525615463033, 0.490201191278175, 0.224027862073854, 0.248289237264544, 0.626155903097242, 0.94530998962, 0.169512212276459, 0.720389195950702, 0.0505251055583358, 0.698552618734539, 0.0339478924870491, 0.982210551621392, 0.661877296399325, 0.823574548587203, 0.961648115655407, 0.727967636892572, 0.675239043310285, 0.524135486688465, 0.690584864933044, 0.923058917280287, 0.279925497947261, 0.592582806246355, 0.189800423802808, 0.301655714400113, 0.273110378533602, 0.437073116889223, 0.221867945743725, 0.750028687762097, 0.659160975134, 0.297228699084371, 0.899841983336955, 0.940142525592819, 0.826975743286312, 0.906821864191443, 0.699459408177063, 0.643372413003817, 0.414709910750389, 0.338854151312262, 0.392046428285539, 0.432181254029274, 0.799300306709483, 0.53010364016518, 0.83987411018461, 0.846017446136102, 0.966542318696156, 0.0224130703136325, 0.0951083095278591, 0.340156800113618, 0.913700114237145, 0.909702165052295, 0.880291021661833, 0.369841040577739, 0.566673917463049, 0.129904052475467, 0.225838124286383, 0.79474611976184, 0.757097277790308, 0.182281930698082, 0.43683678121306, 0.408688043244183, 0.342767896130681, 0.0410993609111756, 0.122907303040847, 0.774846513522789, 0.821215680101886, 0.917958947597072, 0.324944692198187, 0.82686712150462, 0.890137626323849, 0.976657765218988, 0.802807163447142, 0.620720978593454, 0.148597835563123, 0.322460866998881, 0.257811954943463, 0.89846894540824, 0.282446559751406, 0.541350317420438, 0.383074907353148, 0.408896079519764, 0.361857869196683, 0.757965255761519, 0.76706142979674, 0.336052244063467, 0.309799742652103, 0.498503157868981, 0.26244005584158, 0.758614892372862, 0.696643982781097, 0.691610855981708, 0.630582329584286, 0.159666654653847, 0.68286720267497, 0.124332716222852, 0.325191718991846, 0.772778836777434, 0.264704548753798, 0.0900710683781654, 0.0870671425946057, 0.0861371108330786, 0.940812170505524, 0.746780821820721, 0.108701401622966, 0.406559300143272, 0.633501018164679, 0.100139469141141, 0.572460912633687, 0.498886796180159, 0.305858348729089, 0.438014173181728, 0.64346735086292, 0.769598985090852, 0.614872467704117, 0.309185550780967, 0.179521031212062, 0.367358251707628, 0.628589344443753, 0.49061342375353, 0.755762802902609, 0.547388634178787, 0.960398148512468, 0.70118619967252, 0.369968431070447, 0.530221611494198, 0.833572824718431, 0.872962410794571, 0.767327276756987, 0.640161148272455, 0.451970265246928, 0.00181419076398015, 0.201756584690884, 0.209134049713612, 0.691020822618157, 0.792960018618032, 0.069020536262542, 0.758707412518561, 0.934562872396782, 0.398023192537948, 0.944869073107839, 0.563131759176031, 0.877125600352883, 0.513919758843258, 0.767226757714525, 0.032762325135991, 0.794067496201023, 0.260578161105514, 0.504924118518829, 0.588228550041094, 0.466003684792668, 0.81971720373258, 0.543078912654892, 0.629098681733012, 0.699229998979717, 0.0611920461524278, 0.420707281213254, 0.532326074549928, 0.626363403862342, 0.61210541636683, 0.0725685246288776, 0.240231205476448, 0.694840708049014, 0.924698405200616, 0.214044637978077, 0.996561586391181, 0.124272054294124, 0.0936213871464133, 0.96816104860045, 0.167593930149451, 0.831705676624551, 0.153196466621011, 0.928032510448247, 0.212478403467685, 0.619224507827312, 0.48792030243203, 0.97392443404533, 0.876128169475123, 0.739187327679247, 0.0726644534152001, 0.952844584127888, 0.079956975299865, 0.227144148666412, 0.617036697920412, 0.0433553503826261, 0.885906779672951, 0.777480255812407, 0.953097234247252, 0.463224284816533, 0.756394936935976, 0.647401958703995, 0.689930646913126, 0.915713680675253, 0.359993677586317, 0.309201016323641, 0.80172406299971, 0.37825250858441, 0.413076021475717, 0.369931056397036, 0.158351691439748, 0.466690339613706, 0.675279052462429, 0.839463343843818, 0.71888426085934, 0.489989216672257, 0.557963519822806, 0.889969984767959, 0.780990546802059, 0.508734585018829, 0.37386948056519, 0.0268320075701922, 0.54687313362956, 0.509234609548002, 0.662775960285217, 0.94805675232783, 0.426495898980647, 0.100990410195664, 0.630084829637781, 0.73941506491974, 0.489152757450938, 0.822134862886742, 0.555331603623927, 0.792370090261102, 0.189197142608464, 0.218134594848379, 0.139935725368559, 0.0699107849504799, 0.0320600389968604, 0.198781192069873, 0.803292577154934, 0.177155595039949, 0.103483430342749, 0.969729015603662, 0.358960676705465, 0.227047105785459, 0.146968378685415, 0.0309213774744421, 0.78525056457147, 0.896712565096095, 0.978423597291112, 0.765962289646268, 0.788954488700256, 0.798838549293578, 0.199572000419721, 0.642646569525823, 0.446909327525645, 0.853388682939112, 0.49774386594072, 0.826179770752788, 0.92498411424458, 0.740799580235034, 0.853878647321835, 0.516038969391957, 0.922544928034768, 0.578070191899315, 0.0125854355283082, 0.93653458612971, 0.689937360584736, 0.191820221021771, 0.980002827011049, 0.00612950371578336, 0.152231254149228, 0.753500511636958, 0.710635718889534, 0.464286747155711, 0.539280616212636, 0.936696548713371, 0.919709196081385, 0.531410003546625, 0.668291517999023, 0.8565278199967, 0.394768505590037, 0.170387696940452, 0.223712048027664, 0.962624461855739, 0.34642247762531, 0.596265183994547, 0.432610981632024, 0.43734039645642, 0.849443471059203, 0.36918175406754, 0.8357482268475, 0.211098862113431, 0.370749968104064, 0.902853957144544, 0.80517187807709, 0.18266272591427, 0.99213755573146, 0.0350759048014879, 0.217770261224359, 0.491625168826431, 0.389829113380983, 0.635747944703326, 0.0412599672563374, 0.495524066034704, 0.275978177785873, 0.699699725955725, 0.00253413384780288, 0.525128159904853, 0.874406781047583, 0.501205187756568, 0.957805963931605, 0.113309442996979, 0.019440324511379, 0.213787883520126, 0.847250934690237, 0.682643959997222, 0.304549229796976, 0.875393395777792, 0.988973504630849, 0.0311967481393367, 0.718666475266218, 0.685851659392938, 0.685331372777, 0.323351654456928, 0.653485542163253, 0.0566961003933102, 0.714280036045238, 0.799707410158589, 0.321281016105786, 0.552510882262141, 0.567833814769983, 0.226559046190232, 0.0161597041878849, 0.103133757365867, 0.525161945493892, 0.558864328544587, 0.448454127414152, 0.068794475402683, 0.800191269954666, 0.622347963508219, 0.630034809699282, 0.686500137438998, 0.196779183577746, 0.954995960230008, 0.823684701696038, 0.777487255632877, 0.991282063303515, 0.749071105848998, 0.221222479362041, 0.0104664508253336, 0.94431388610974, 0.42660741135478, 0.231606764020398, 0.858432770008221, 0.860882739303634, 0.34367073699832, 0.566733361920342, 0.276100607356057, 0.807627876289189, 0.0954804811626673, 0.659552552504465, 0.340250159613788, 0.584376116050407, 0.752787398640066, 0.444549790816382, 0.573426559567451, 0.313154024537653, 0.583494216669351, 0.483133491128683, 0.489511491032317, 0.611506707966328, 0.115509082330391, 0.860434833914042, 0.986133378231898, 0.520857575116679, 0.112023608060554, 0.649017740506679, 0.0642707417719066, 0.964710023254156, 0.473144950577989, 0.988888219930232, 0.867568374844268, 0.424685476580635, 0.535276068840176, 0.00368020194582641, 0.492363908560947, 0.22955672792159, 0.0105552098248154, 0.132949529681355, 0.262833642307669, 0.789901982992887, 0.418283185688779, 0.0599505722057074, 0.688781585777178, 0.348022612277418, 0.0022046088706702, 0.359848840394989, 0.96220533340238, 0.26195351453498, 0.648078184109181, 0.854865761240944, 0.976172536611557, 0.21778170671314, 0.366131584392861, 0.411253718892112, 0.347585044102743, 0.861726809525862, 0.793758688727394, 0.561874046456069, 0.142363130114973, 0.995368130970746, 0.610652939183637, 0.774910867912695, 0.180390730965883, 0.207898023538291, 0.210500702960417, 0.122632145183161, 0.963247831445187, 0.229282746557146, 0.00899550365284085, 0.743316345615312, 0.411071793874726, 0.649562026839703, 0.935616058995947, 0.512678687926382, 0.781916251871735, 0.202930553816259, 0.440406591165811, 0.835982216522098, 0.588324965443462, 0.707541161216795, 0.216485942713916, 0.337843096116558, 0.687637854600325, 0.694892174797133, 0.902814090717584, 0.120111820753664, 0.581117763649672, 0.994154472835362, 0.0553591488860548, 0.107527822488919, 0.0303732336033136, 0.407859860686585, 0.905584805179387, 0.441940115764737, 0.50978111452423, 0.97273276746273, 0.882490286137909, 0.287773457588628, 0.171000141650438, 0.786767660873011, 0.253843790851533, 0.255465920083225, 0.308531409827992, 0.477953705238178, 0.0519169725012034, 0.540956609416753, 0.154846774414182, 0.383960497099906, 0.861564192920923, 0.00752651900984347, 0.523947555106133, 0.614303425187245, 0.452583259670064, 0.803628797642887, 0.370584713062271, 0.755196412559599, 0.552135017002001, 0.0151524152606726, 0.809359599137679, 0.75017999089323, 0.403430354548618, 0.742492314660922, 0.0407378482632339, 0.586590654216707, 0.370698592159897, 0.998476498760283, 0.907150311395526, 0.584493603091687, 0.215629647485912, 0.154198286822066, 0.187613315414637, 0.763280612416565, 0.979748753830791, 0.919624150497839, 0.190033343154937, 0.709146238863468, 0.319378637010232, 0.63226525648497, 0.113028575433418, 0.784492762759328, 0.841583132045344, 0.58860769122839, 0.365569554734975, 0.912733611650765, 0.40532320481725, 0.469806988257915, 0.740595693467185, 0.649585058912635, 0.551680438453332, 0.803546288749203, 0.266639578621835, 0.72000957489945, 0.784368597436696, 0.682293208315969, 0.846919790608808, 0.762568810023367, 0.307593358214945, 0.582655605627224, 0.207152150338516, 0.192923614289612, 0.359944900497794, 0.229508531745523, 0.447255559498444, 0.469163330970332, 0.51522917766124, 0.710733385989442, 0.917870621196926, 0.724946552654728, 0.814591881120577, 0.728415706660599, 0.368014480685815, 0.167438768548891, 0.208689883351326, 0.0597604932263494, 0.493736804230139, 0.343189745442942, 0.566354803973809, 0.158048330573365, 0.371554979123175, 0.707275984808803, 0.80879473942332, 0.305027137044817, 0.21863532345742, 0.940836825873703, 0.0891549456864595, 0.471684027463198, 0.114925465080887, 0.359661899972707, 0.561475561233237, 0.438617490930483, 0.742599151097238, 0.59793760557659, 0.547305121552199, 0.186299775494263, 0.328532898565754, 0.907397585222498, 0.0570698576048017, 0.871221818029881, 0.812638330971822, 0.167275724932551, 0.604715018766001, 0.719323352212086, 0.146683736937121, 0.030764076160267, 0.965944251045585, 0.484099846333265, 0.896890627918765, 0.493849067483097, 0.949982817051932, 0.475978490198031, 0.855085253017023, 0.548597631976008, 0.677511579589918, 0.800946189556271, 0.212203885661438, 0.481530643068254, 0.67040012544021, 0.380454719532281, 0.638210525037721, 0.831939764553681, 0.75791493896395, 0.0163347593043, 0.874640173278749, 0.978919329121709, 0.0279686618596315, 0.603906699456275, 0.325781769119203, 0.83012081659399, 0.87641319562681, 0.566293722484261, 0.443134146975353, 0.643166835885495, 0.50316036096774, 0.954935207497329, 0.0495021708775312, 0.71875398978591, 0.458132177125663, 0.284932358656079, 0.833258487982675, 0.593742285855114, 0.232698033563793, 0.370992517098784, 0.126291529042646, 0.419836586108431, 0.0550733113195747, 0.193297164049, 0.389814133523032, 0.0408909153193235, 0.119742008857429, 0.0963508477434516, 0.320356306852773, 0.0628326598089188, 0.776118252892047, 0.204186249524355, 0.297879328019917, 0.270763511070982, 0.332327314186841, 0.379567924654111, 0.957255876855925, 0.0605127364397049, 0.845707227941602, 0.524631295353174, 0.325755237834528, 0.110937610268593, 0.75218552001752, 0.260240651434287, 0.120974946534261, 0.470583563204855, 0.0679423115216196, 0.709851996041834, 0.640370265115052, 0.760155219817534, 0.490326981525868, 0.714379087556154, 0.279203770915046, 0.437578638317063, 0.637385812820867, 0.45179330999963, 0.0323167690075934, 0.503172723809257, 0.134268457302824, 0.0687409022357315, 0.327924631768838, 0.659112301422283, 0.437960103387013, 0.763784825801849, 0.969024687772617, 0.396040805615485, 0.511843671789393, 0.112226026831195, 0.2919939563144, 0.386387607781217, 0.627518968656659, 0.486018694471568, 0.636940563097596, 0.64090036530979, 0.305422897217795, 0.170779120177031, 0.275626598158851, 0.0399546283297241, 0.530222168425098, 0.0637374625075608, 0.15574282896705, 0.0386561409104615, 0.64832698716782, 0.548114001052454, 0.716736173024401, 0.913377426099032, 0.462437190115452, 0.642535324208438, 0.563876203494146, 0.333562181564048, 0.314121963921934, 0.360791054321453, 0.771765416720882, 0.928775905631483, 0.798027148935944, 0.564099924406037, 0.588276706635952, 0.387102958746254, 0.538517222972587, 0.241420002887025, 0.833665811922401, 0.714759069029242, 0.975525598041713, 0.0245571427512914, 0.729954986833036, 0.360193769214675, 0.0994546189904213, 0.74384393892251, 0.929154738783836, 0.0016889963299036, 0.604328516870737, 0.780568923801184, 0.277538565220311, 0.999487799592316, 0.155518087325618, 0.375074702082202, 0.216007667826489, 0.855680725770071, 0.74545754538849, 0.15285931667313, 0.169231446227059, 0.958105371566489, 0.919468614272773, 0.550246902275831, 0.453199676703662, 0.734624997712672, 0.603131785755977, 0.814369122264907, 0.610151410102844, 0.28671961161308, 0.377699196804315, 0.645668612094596, 0.459602150600404, 0.0636904251296073, 0.730845224345103, 0.787010733736679, 0.775095997145399, 0.963043381925672, 0.781875079963356, 0.165395512711257, 0.102167158154771, 0.112617174163461, 0.236617875285447, 0.620523048797622, 0.211341302841902, 0.019465038087219, 0.0728287426754832, 0.768831913825125, 0.170224415138364, 0.196226181462407, 0.932493109954521, 0.909011497627944, 0.731983660254627, 0.502005791990086, 0.129261485766619, 0.437228534370661, 0.927112497389317, 0.492754122940823, 0.229928021784872, 0.768724255962297, 0.596535144606605, 0.287562065990642, 0.607624302618206, 0.0150145837105811, 0.465425665490329, 0.752262640045956, 0.234585015103221, 0.28010435262695, 0.689670686610043, 0.729054742958397, 0.54628744465299, 0.567548515507951, 0.0919751517940313, 0.978125780122355, 0.815651939250529, 0.414205363020301, 0.580356651451439, 0.315433646552265, 0.0944330187048763, 0.278594660805538, 0.641516709001735, 0.942213162546977, 0.963663939386606, 0.117009649984539, 0.16886896174401, 0.770162808941677, 0.0290391549933702, 0.509347400860861, 0.154737452976406, 0.218915889505297, 0.81278143147938, 0.7918132818304, 0.361911590676755, 0.533447762485594, 0.205716119380668, 0.145926505327225, 0.567516615381464, 0.661405420862138, 0.912792915478349, 0.812409560661763, 0.166139237815514, 0.912690614815801, 0.60477176355198, 0.303903713356704, 0.175798091338947, 0.779296505032107, 0.868319496512413, 0.902708767447621, 0.16225887555629, 0.171088035451248, 0.955510934116319, 0.733020378276706, 0.343892005970702, 0.397802587365732, 0.169556392123923, 0.718375637196004, 0.651486968854442, 0.704436842352152, 0.171062720939517, 0.727351875510067, 0.82597722648643, 0.450209181522951, 0.64966600574553, 0.559782944153994, 0.692732505267486, 0.327601307071745, 0.104534781305119, 0.736263470491394, 0.734104784205556, 0.127375549869612, 0.0714721316471696, 0.757345570018515, 0.544909867458045, 0.874999487074092, 0.492404003627598, 0.948826359352097, 0.613431945443153, 0.462209581630304, 0.51792767830193, 0.114373931894079, 0.414908068953082, 0.44810791220516, 0.53870382742025, 0.391342780552804, 0.0629779412411153, 0.0342692008707672, 0.674611888825893, 0.972176955081522, 0.172799657098949, 0.106380753451958, 0.351197320735082, 0.17712702229619, 0.688318250002339, 0.327085983240977, 0.200052581494674, 0.735188523074612, 0.706861588172615, 0.0897164207417518, 0.272563348524272, 0.0622873618267477, 0.498308047885075, 0.913953924551606, 0.287869580555707, 0.947207218501717, 0.860370331211016, 0.742887367028743, 0.515208676923066, 0.316783347865567, 0.0358943003229797, 0.954185078153387, 0.787259324686602, 0.51324513158761, 0.0509291752241552, 0.471745676593855, 0.469335162546486, 0.188746983418241, 0.620796443894506, 0.457652227021754, 0.039797778474167, 0.775200942996889, 0.705058598192409, 0.363446604926139, 0.856544058537111, 0.587862790795043, 0.705456369556487, 0.71006122790277, 0.0929951064754277, 0.99938575970009, 0.456101974938065, 0.930131077067927, 0.0755331730470061, 0.787744448520243, 0.784907332621515, 0.922089783241972, 0.0862829545512795, 0.667868304532021, 0.510595293482766, 0.867293874267489, 0.566685023484752, 0.30555938417092, 0.718691570218652, 0.32572992821224, 0.552464926848188, 0.408631222788244, 0.287290554260835, 0.636740854475647, 0.764731423929334, 0.412437052000314, 0.872906443662941, 0.059664097847417, 0.278291850583628, 0.00475005409680307, 0.881868486525491, 0.091316367033869, 0.528467135038227, 0.829737195046619, 0.946491332026199, 0.362233009189367, 0.92408715817146, 0.754694551695138, 0.525746026542038, 0.561734290095046, 0.0797179362270981, 0.784518458880484, 0.735162960132584, 0.219211782328784, 0.951858182437718, 0.443051582900807, 0.3887527838815, 0.718419762095436, 0.578670738264918, 0.425928557757288, 0.020913421176374, 0.863156679552048), actual = c(0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.418269178830087, 0.0319914394058287, 0.83267064881511, 0.120048529934138, 0.850769061595201, 0.704163889866322, 0.688647508388385, 0.388169376412407, 0.235873518511653, 0.305275430204347, 0.704544107662514, 0.0736174308694899, 0.00158441136591136, 0.574607048649341, 0.358301207888871, 0.426050873706117, 0.813968153903261, 0.831767019350082, 0.435156848048791, 0.266379806911573, 0.231182418065146, 0.251312179490924, 0.73383415187709, 0.146874882513657, 0.960093305911869, 0.651991697493941, 0.917773625114933, 0.110305027803406, 0.227955954615027, 0.129370386945084, 0.135677440091968, 0.864823669893667, 0.643865901045501, 0.799078318523243, 0.0288647210691124, 0.839512028731406, 0.812747152289376, 0.029834404354915, 0.700205936096609, 0.663301778957248, 0.170095163397491, 0.290860004723072, 0.38259712443687, 0.850653278641403, 0.726265746867284, 0.629671757575125, 0.363080410752445, 0.98313592444174, 0.663945958251134, 0.207980112172663, 0.276151685509831, 0.150370533578098, 0.505911645712331, 0.735985823674127, 0.591494632419199, 0.816139286849648, 0.460804444272071, 0.199037858750671, 0.643401654670015, 0.174142281292006, 0.554434739751741, 0.751914515858516, 0.049613511422649, 0.16643446078524, 0.712931485148147, 0.580154519760981, 0.557068620808423, 0.61682896129787, 0.285756537690759, 0.559460129588842, 0.95983043895103, 0.058694644831121, 0.382214896380901, 0.698921484174207, 0.502311895601451, 0.265141472220421, 0.400469424435869, 0.591164060635492, 0.283664986258373, 0.962934354320168, 0.703196813818067, 0.736970419995487, 0.854116367409006, 0.945274621248245, 0.538409769069403, 0.686233749613166, 0.952042227843776, 0.914517075987533, 0.253524907398969, 0.229877654695883, 0.0413119895383716, 0.910809444496408, 0.109716954641044, 0.130084289237857, 0.180537000531331, 0.106489816214889, 0.852498036576435, 0.692625232273713, 0.514627608470619, 0.461202511098236, 0.114728576503694, 0.556871064007282, 0.681678545661271, 0.625959190540016, 0.581016406882554, 0.0546654553618282, 0.384929123567417, 0.293911927612498, 0.126299536554143, 0.312126314966008, 0.105855247471482, 0.238227296620607, 0.269790570484474, 0.372535414295271, 0.375663021113724, 0.40854539652355, 0.676275110803545, 0.939417948247865, 0.0194915330503136, 0.379078320926055, 0.888956650625914, 0.197532723192126, 0.495913089485839, 0.911691097542644, 0.852846344001591, 0.0770766341593117, 0.18837206531316, 0.354866301640868, 0.413748137187213, 0.8698206897825, 0.360580585896969, 0.935069374041632, 0.0859970881137997, 0.769144658464938, 0.658942776499316, 0.0356038494501263, 0.211600447306409, 0.728827298618853, 0.945508699398488, 0.806500338949263, 0.912992738420144, 0.947635015007108, 0.879836038220674, 0.926173895830289, 0.250892221461982, 0.395859670126811, 0.904532209504396, 0.0664109918288887, 0.706218476174399, 0.401788196992129, 0.0761853430885822, 0.169728833716363, 0.416721664369106, 0.585426335223019, 0.938211173284799, 0.490791241638362, 0.953907246002927, 0.32725056190975, 0.0203147258143872, 0.96800549607724, 0.840805055573583, 0.181480299215764, 0.452811654191464, 0.142926789820194, 0.137962429085746, 0.59021999919787, 0.834704342065379, 0.479884079424664, 0.376280402531847, 0.00806994852609932, 0.889901407528669, 0.328724502120167, 0.032007894013077, 0.715617540990934, 0.135196719551459, 0.723659364040941, 0.0228167183231562, 0.169911758508533, 0.336876734159887, 0.162235420197248, 0.191189534496516, 0.123103210702538, 0.344139357097447, 0.019775423919782, 0.769445874728262, 0.908411390148103, 0.745942013338208, 0.931060454109684, 0.288685251260176, 0.229760462185368, 0.485229942481965, 0.249188120476902, 0.0765731467399746, 0.738890297478065, 0.291169802192599, 0.691428798949346, 0.496654588263482, 0.764744749525562, 0.409030145965517, 0.119441584683955, 0.988673399202526, 0.445244303671643, 0.912035796791315, 0.849292309489101, 0.0678791829850525, 0.69153696927242, 0.342330717481673, 0.368873494910076, 0.922020060010254, 0.80037244129926, 0.374734464334324, 0.788096035830677, 0.264158073812723, 0.438283235998824, 0.68139428156428, 0.856517600361258, 0.0546802901662886, 0.714336409466341, 0.518271129112691, 0.890344155719504, 0.598879103315994, 0.576018112478778, 0.323638780042529, 0.450505547691137, 0.275514014530927, 0.0197572004981339, 0.229404487647116, 0.728745330357924, 0.143374274019152, 0.188713739160448, 0.905786803225055, 0.422285490203649, 0.168450549244881, 0.316382957389578, 0.137598573230207, 0.504591479199007, 0.210352129302919, 0.181950777303427, 0.712159992661327, 0.226978817256168, 0.747519452590495, 0.865318130934611, 0.253815383883193, 0.270079816924408, 0.9997291311156, 0.0734566543251276, 0.801493625622243, 0.200423094909638, 0.128738986561075, 0.334881704766303, 0.39375311578624, 0.963482315652072, 0.314643652411178, 0.704963883617893, 0.575569820357487, 0.9229252403602, 0.580415569711477, 0.167808043304831, 0.35028651310131, 0.466664479579777, 0.676462211180478, 0.979673165129498, 0.0684381085447967, 0.917912476463243, 0.469483912456781, 0.130387199809775, 0.775438884040341, 0.862940185703337, 0.683968924218789, 0.535308368736878, 0.423670972231776, 0.0761426382232457, 0.0151021755300462, 0.394007671391591, 0.889606654644012, 0.710825817426667, 0.73271277686581, 0.177044598385692, 0.96170574054122, 0.37837937916629, 0.0129664605483413, 0.110413758317009, 0.928190688602626, 0.855609963648021, 0.753569402731955, 0.562430997611955, 0.270214067073539, 0.422005004016683, 0.985337452264503, 0.392584259621799, 0.396004476118833, 0.0155183358583599, 0.254872597521171, 0.128658800851554, 0.490845843683928, 0.877164500765502, 0.646212404360995, 0.351577325491235, 0.313726936466992, 0.660851576831192, 0.387806548504159, 0.270968031371012, 0.67154369992204, 0.606348267756402, 0.350982166826725, 0.140139335533604, 0.210118220886216, 0.934970948146656, 0.380819906713441, 0.481133360648528, 0.204436824889854, 0.248603750485927, 0.413210913306102, 0.315081601263955, 0.827873297734186, 0.0973491626791656, 0.442575828172266, 0.601358298677951, 0.438448785571381, 0.805679996265098, 0.0675713028758764, 0.510812958236784, 0.195242096902803, 0.242996698478237, 0.5241292596329, 0.884698046604171, 0.67129739234224, 0.0833262007217854, 0.627473836299032, 0.599125961074606, 0.786331441719085, 0.619620590936393, 0.905970202060416, 0.954685209086165, 0.927329009398818, 0.859871272696182, 0.76303736702539, 0.291237673023716, 0.817213410977274, 0.876495853764936, 0.0589781636372209, 0.328889321302995, 0.338624604279175, 0.00548872910439968, 0.493117195786908, 0.388027440290898, 0.651536278426647, 0.590614925837144, 0.717905124649405, 0.902860344620422, 0.96111020934768, 0.577948126476258, 0.383845523698255, 0.00174208986572921, 0.274669737787917, 0.971432662103325, 0.0428223656490445, 0.241046502953395, 0.0474115968681872, 0.0594358542002738, 0.525420589838177, 0.108147477963939, 0.330720196478069, 0.863993540871888, 0.727826320799068, 0.870464926818386, 0.619030814152211, 0.663749976316467, 0.566105401609093, 0.497873553773388, 0.761005565291271, 0.214021624298766, 0.634892113273963, 0.86786852707155, 0.353785441024229, 0.579208356561139, 0.615470022661611, 0.431089972145855, 0.915112296119332, 0.72469742805697, 0.635741601698101, 0.651474768761545, 0.038441423792392, 0.0814034808427095, 0.457995076663792, 0.295514284865931, 0.882821693317965, 0.149371547624469, 0.787768140435219, 0.340731230797246, 0.286796142812818, 0.188066242728382, 0.528767424402758, 0.740465189097449, 0.686166308121756, 0.382497127866372, 0.832770321751013, 0.0236812476068735, 0.463750352151692, 0.110363762360066, 0.937157639767975, 0.0982943179551512, 0.318947084480897, 0.491861020447686, 0.710163952084258, 0.0410647871904075, 0.472729665227234, 0.371622791048139, 0.681224781321362, 0.142251541372389, 0.740758349886164, 0.816866745240986, 0.0774726064410061, 0.410505956504494, 0.717781321844086, 0.55896884133108, 0.334404509747401, 0.128737088991329, 0.559680146398023, 0.701840206747875, 0.934059850173071, 0.383665405912325, 0.650890186894685, 0.183402249123901, 0.906569159822538, 0.216196369379759, 0.634159766836092, 0.681435817852616, 0.80837809224613, 0.0794811842497438, 0.114190601278096, 0.313419060781598, 0.350824763299897, 0.936716339318082, 0.843572212848812, 0.995977975195274, 0.0916275435592979, 0.242994049796835, 0.806492044124752, 0.647742439759895, 0.980467052431777, 0.799343309365213, 0.887917852727696, 0.372197635471821, 0.209598089102656, 0.776590142631903, 0.550199217395857, 0.982313635759056, 0.266897213645279, 0.758306016214192, 0.975578851765022, 0.808373471256346, 0.234851964516565, 0.159891433315352, 0.143862745258957, 0.485425197519362, 0.657496273051947, 0.121260631131008, 0.943030499387532, 0.346751067321748, 0.708278339821845, 0.775815451750532, 0.793533832766116, 0.0995440031401813, 0.604993206448853, 0.614694219548255, 0.531313065439463, 0.286114409798756, 0.0206715897656977, 0.586465077009052, 0.416893856134266, 0.367378443479538, 0.239686659071594, 0.945648810360581, 0.64820157061331, 0.236510147806257, 0.947570658987388, 0.537809033179656, 0.0328426449559629, 0.443261799868196, 0.0423302571289241, 0.791226055938751, 0.238728614989668, 0.466236889129505, 0.628168272320181, 0.750340417493135, 0.0808479771949351, 0.31269977777265, 0.100286081666127, 0.753567538224161, 0.0135961908381432, 0.756214524619281, 0.46248400141485, 0.752727854764089, 0.323431108845398, 0.89632757124491, 0.327173450496048, 0.569025457603857, 0.715934713603929, 0.541717837797478, 0.0250117329414934, 0.0289497671183199, 0.0616373459342867, 0.341314273187891, 0.972636017249897, 0.624089404940605, 0.788378562545404, 0.150786260608584, 0.621112938504666, 0.88176778354682, 0.800744945416227, 0.588500433601439, 0.261378840077668, 0.163062537787482, 0.737124521750957, 0.722089847316965, 0.0392824090085924, 0.0849826382473111, 0.682740413816646, 0.327115298248827, 0.581807010574266, 0.296033587306738, 0.688550368417054, 0.298443488543853, 0.282135835383087, 0.370148967951536, 0.22967780684121, 0.430998813360929, 0.729679262964055, 0.0761185812298208, 0.207773074740544, 0.998653683811426, 0.2281751295086, 0.873271159362048, 0.00180951203219593, 0.893216219497845, 0.683686406118795, 0.489210573490709, 0.729199666064233, 0.8112842743285, 0.447132449829951, 0.312885309336707, 0.526051118969917, 0.886908279499039, 0.0465539037249982, 0.378143820445985, 0.0134347379207611, 0.911999673582613, 0.502231606049463, 0.478242291370407, 0.370894004125148, 0.33285696269013, 0.713682365836576, 0.174169628880918, 0.945568354800344, 0.443523678695783, 0.0109428407158703, 0.754356834106147, 0.462885126005858, 0.416649802122265, 0.0685499359387904, 0.94670404237695, 0.214171580737457, 0.790222581475973, 0.261061083525419, 0.0233128126710653, 0.784664409700781, 0.658912977669388, 0.347069502109662, 0.68981596082449, 0.170880855526775, 0.914676046697423, 0.476317071821541, 0.606237317435443, 0.863338177790865, 0.162896012654528, 0.506814738269895, 0.785948827629909, 0.741330499993637, 0.171082610730082, 0.566873837029561, 0.0452805499080569, 0.877727316459641, 0.613322135293856, 0.921164075611159, 0.757342973025516, 0.546008001314476, 0.869095561560243, 0.925404828041792, 0.47999326325953, 0.398211271269247, 0.0340583571232855, 0.564978779992089, 0.377725796308368, 0.031399687519297, 0.132116650929675, 0.732628198340535, 0.404282547067851, 0.762403355445713, 0.534465890610591, 0.116430586436763, 0.0697410313878208, 0.0163250938057899, 0.486001096665859, 0.0463410725351423, 0.857629691483453, 0.359172602649778, 0.242716274224222, 0.250924784922972, 0.934901285218075, 0.394745143363252, 0.869469766737893, 0.404122419655323, 0.743887844029814, 0.121871591778472, 0.788006282178685, 0.681756646838039, 0.0572632371913642, 0.056838937336579, 0.829458102351055, 0.3404957188759, 0.205887888558209, 0.100890171481296, 0.992913490626961, 0.173811485758051, 0.194799395976588, 0.355936521664262, 0.110711997374892, 0.782912540948018, 0.637335670180619, 0.0748554009478539, 0.940703390864655, 0.145346406148747, 0.448301453376189, 0.617130917962641, 0.0790507269557565, 0.718551761005074, 0.106928675901145, 0.189817223697901, 0.63816084060818, 0.933286145562306, 0.523483583005145, 0.396364086307585, 0.909357528667897, 0.358597348211333, 0.842533944640309, 0.260680858278647, 0.460564179578796, 0.788581179222092, 0.595492229796946, 0.690753849921748, 0.3320548532065, 0.0846005119383335, 0.985918562393636, 0.52618767041713, 0.667999957688153, 0.949844015529379, 0.194025782635435, 0.914241702063009, 0.881462770979851, 0.904018108034506, 0.972060311120003, 0.184982291189954, 0.108519562752917, 0.906513973604888, 0.435997929889709, 0.742295665666461, 0.655078710289672, 0.468699970748276, 0.106709477957338, 0.844471658812836, 0.526229350594804, 0.672216574428603, 0.4583090164233, 0.373961345525458, 0.491120818303898, 0.745078021194786, 0.760730840265751, 0.614116100361571, 0.348143086768687, 0.870892083039507, 0.160406570881605, 0.0990889903623611, 0.668824085965753, 0.481053998228163, 0.398856307612732, 0.88484158902429, 0.821260419208556, 0.677653227001429, 0.883628463605419, 0.694820312317461, 0.497633184306324, 0.668367308797315, 0.103284625103697, 0.37907336326316, 0.978367952397093, 0.579098121728748, 0.487761456985027, 0.351240271003917, 0.0495787998661399, 0.470340981613845, 0.224677375284955, 0.118935908190906, 0.680681633064523, 0.507994971005246, 0.801607610657811, 0.784186670323834, 0.302226765314117, 0.0900462742429227, 0.565139553742483, 0.871667765313759, 0.329592834226787, 0.42189856339246, 0.864652083255351, 0.172858366044238, 0.289632493397221, 0.974650426302105, 0.603246439481154, 0.803728983271867, 0.597462091362104, 0.96440812246874, 0.892530472716317, 0.775553562212735, 0.539480148814619, 0.0804681198205799, 0.227489920333028, 0.575828186469153, 0.224933906458318, 0.485234023770317, 0.97202310222201, 0.746384980389848, 0.143145284615457, 0.481448089936748, 0.241779887350276, 0.545236818026751, 0.629100117832422, 0.638325425097719, 0.563527408987284, 0.689344014739618, 0.932789409765974, 0.0166797174606472, 0.904902479844168, 0.68932510796003, 0.320852477103472, 0.721967702498659, 0.607515896204859, 0.264320081099868, 0.407790969824418, 0.118117382749915, 0.969335896661505, 0.53510178043507, 0.476486336672679, 0.767449918435887, 0.411129513755441, 0.0869310940615833, 0.0898190443404019, 0.228187305852771, 0.13961966894567, 0.794741614488885, 0.234765800414607, 0.132765103597194, 0.499119214713573, 0.187572889961302, 0.811306262621656, 0.731547240167856, 0.00213845283724368, 0.138637129683048, 0.819525072118267, 0.00640985346399248, 0.0891809349413961, 0.280869877431542, 0.643010459374636, 0.131442243466154, 0.0391038281377405, 0.547441220376641, 0.885268375976011, 0.255633506691083, 0.11105227121152, 0.55794205958955, 0.701394140953198, 0.421574321575463, 0.132413394516334, 0.537146751303226, 0.0966174511704594, 0.0929652098566294, 0.773761730175465, 0.247985112480819, 0.383543754694983, 0.0537740001454949, 0.42214417969808, 0.89089847006835, 0.463661574060097, 0.833771396428347, 0.769522778689861, 0.464185071177781, 0.25478923949413, 0.431079431204125, 0.836793584050611, 0.736304600723088, 0.555888622067869, 0.348446182906628, 0.446951870806515, 0.453392629977316, 0.379418948898092, 0.687344946665689, 0.858140103053302, 0.0788010486867279, 0.877070867456496, 0.229102015960962, 0.894335099495947, 0.437148400815204, 0.857466156361625, 0.747559168608859, 0.644180872943252, 0.105267082108185, 0.825359328882769, 0.873142770025879, 0.986844754312187, 0.991919439518824, 0.577953821048141, 0.659239220898598, 0.581124349497259, 0.730493803275749, 0.968576277140528, 0.468773652566597, 0.353149529546499, 0.571405422640964, 0.252452804706991, 0.935189789393917, 0.513669353444129, 0.348343615187332, 0.877056606579572, 0.640353857306764, 0.516680004308, 0.667565040988848, 0.944150522351265, 0.911769870435819, 0.17867725645192, 0.870455175871029, 0.97739082807675, 0.876430245582014, 0.00684812339022756, 0.727862955536693, 0.228493652073666, 0.67094328510575, 0.303151779575273, 0.0905410021077842, 0.737323182169348, 0.657039669342339, 0.174851401476189, 0.518131703371182, 0.80423296848312, 0.122283694101498, 0.637877129949629, 0.788477750727907, 0.827021267963573, 0.64363248902373, 0.534443101845682, 0.580800657626241, 0.670707708457485, 0.33740273793228, 0.182252595666796, 0.154037077678367, 0.25761865475215, 0.566847536480054, 0.703319350956008, 0.148341116961092, 0.449390248395503, 0.791169615928084, 0.140557184582576, 0.335106670856476, 0.884931236505508, 0.843419373501092, 0.136099284980446, 0.719111868878826, 0.413166112499312, 0.186076378915459, 0.345612172503024, 0.689810817362741, 0.364141395082697, 0.710192499216646, 0.296642565634102, 0.78857452608645, 0.133346946910024, 0.39655307168141, 0.0203518574126065, 0.798267056699842, 0.977371111512184, 0.715718791121617, 0.811651839874685, 0.995700210798532, 0.727685597259551, 0.439260331913829, 0.636065318714827, 0.539006637176499, 0.242560676997527, 0.926324193831533, 0.453450269298628, 0.919902910711244, 0.791048643644899, 0.854921989375725, 0.0673762769438326, 0.831873598508537, 0.773771414533257, 0.679849501699209, 0.594568065134808, 0.177336567081511, 0.0716384183615446, 0.109734469093382, 0.142953683389351, 0.371251607313752, 0.720279330853373, 0.19714603270404, 0.48206963785924, 0.168891249224544, 0.245586952194571, 0.976130369585007, 0.946119153406471, 0.397902994416654, 0.866146412212402, 0.1839377253782, 0.408557585440576, 0.888489114120603, 0.740056865150109, 0.299589572940022, 0.686868432210758, 0.130031543551013, 0.893705754773691, 0.324845924740657, 0.534134450368583, 0.39730504527688, 0.792438457487151, 0.654052619589493, 0.390240730252117, 0.32164367637597, 0.881873867940158, 0.301675896625966, 0.0793153250124305, 0.847860206849873, 0.438707457156852, 0.504936483921483, 0.872894189087674, 0.681395490653813, 0.558372015599161, 0.307901591993868, 0.803933222778141, 0.555716551840305, 0.393264551181346, 0.345856854924932, 0.286181796109304, 0.646145915146917, 0.978765983833, 0.242789536481723, 0.175291381776333, 0.36106302915141, 0.895454082638025, 0.797190922312438, 0.276026846608147, 0.614729651715606, 0.591676809126511, 0.880802943836898, 0.00914959539659321, 0.269189895130694, 0.500032889423892, 0.523768193321303, 0.834294548956677, 0.466540390858427, 0.902943759225309, 0.802052042679861, 0.331366492900997, 0.331782853929326, 0.351370082003996, 0.925900752656162, 0.396131598157808, 0.57542492938228, 0.284597652032971, 0.450561186531559, 0.437379579525441, 0.672737350920215, 0.60447237896733, 0.69777595391497, 0.686138364952058, 0.877023469423875, 0.0326972138136625, 0.976812795503065, 0.765143192606047, 0.715645940043032, 0.809631214709952, 0.592677973210812, 0.953347146045417, 0.611421049805358, 0.433756177080795, 0.13659440795891, 0.802079878048971, 0.940852921456099, 0.374469339614734, 0.986937748268247, 0.421340489527211, 0.268249290063977, 0.603002201300114), actual = c(1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"