Gets probability thresholds given predicted probabilities, outcomes and NMB.

Gets probability thresholds given predicted probabilities, outcomes and NMB.

Usage

get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)

Arguments

predicted

A vector of predicted probabilities.

actual

A vector of actual outcomes.

nmb

A named vector containing NMB assigned to true positives, true negatives, false positives and false negatives

cutpoint_methods

Which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.

Value

Returns a list.

Examples

# get thresholds using default (all inbuilt) cutpoint methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#>              all             none value_optimising           youden 
#>       0.00000000       1.00000000       0.03387607       0.83239289 
#>  cost_minimising   prod_sens_spec            roc01   index_of_union 
#>       0.50000000       0.45678137       0.45678137       0.47650923 

# get cutpoints using user-defined functions
# These functions must take the \code{predicted} and \code{actual} as arguments
# They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
fx_roc01 <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::minimize_metric,
    metric = cutpointr::roc01,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

fx_sum_sens_spec <- function(predicted, actual, ...) {
  cutpointr::cutpointr(
    x = predicted, class = actual, method = cutpointr::maximize_metric,
    metric = cutpointr::sum_sens_spec,
    silent = TRUE
  )[["optimal_cutpoint"]]
}

get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.103838251437992, 0.635927268536761, 0.391475142445415, 0.240082031581551, 0.249259727075696, 0.540844105184078, 0.00126163265667856, 0.271921451902017, 0.696797877782956, 0.135459222365171, 0.751530053559691, 0.511892548762262, 0.887126338668168, 0.0811601660680026, 0.183809285517782, 0.245910613564774, 0.439262739615515, 0.119331780588254, 0.182415828574449, 0.171166643034667, 0.983831879450008, 0.369243850465864, 0.206653881818056, 0.367296059150249, 0.633256692439318, 0.0472499360330403, 0.816027790075168, 0.566801271867007, 0.0189452592749149, 0.640517632476985, 0.461929213488474, 0.820188688347116, 0.577304312493652, 0.3983897222206, 0.275943249696866, 0.496205626986921, 0.727858843980357, 0.574724032077938, 0.633185527054593, 0.111218478763476, 0.589462105184793, 0.44674955890514, 0.782209360506386, 0.601642461493611, 0.947507843840867, 0.951000145636499, 0.304404590511695, 0.386746449396014, 0.56729997578077, 0.764078400563449, 0.238164430251345, 0.459277987014502, 0.403878055280074, 0.0162610036786646, 0.632416301872581, 0.858992749359459, 0.216072119539604, 0.798390393145382, 0.718745733145624, 0.0378174160141498, 0.928323582047597, 0.00798895000480115, 0.143317632377148, 0.12558118160814, 0.429835373535752, 0.683587834704667, 0.410226817242801, 0.13305686111562, 0.806760808220133, 0.680092916358262, 0.289273463422433, 0.449291446711868, 0.591224441071972, 0.593837106600404, 0.577560784295201, 0.257696118671447, 0.832052277866751, 0.816473334096372, 0.102304717991501, 0.781384170521051, 0.0385500078555197, 0.737981121754274, 0.911369380075485, 0.343085215659812, 0.0711116841994226, 0.577410272788256, 0.214427635539323, 0.848000297322869, 0.387514282483608, 0.391838653711602, 0.202410239959136, 0.0203361411113292, 0.423031881218776, 0.884976272471249, 0.316781450295821, 0.572629194008186, 0.760698886122555, 0.704933046596125, 0.838724289555103, 0.725317864213139, 0.705098729813471, 0.4608969676774, 0.95280605298467, 0.872945562936366, 0.542447498533875, 0.100820912979543, 0.323733031749725, 0.348386453231797, 0.472132832976058, 0.45588583429344, 0.796461144462228, 0.808059542439878, 0.249943102011457, 0.666736723622307, 0.415121147641912, 0.665110024623573, 0.00576953473500907, 0.489100295351818, 0.24240571539849, 0.309757327195257, 0.772228258429095, 0.667488772422075, 0.757517314748839, 0.258105394663289, 0.125359342433512, 0.886635305825621, 0.504416989628226, 0.721637476468459, 0.0147008213680238, 0.787787279346958, 0.315190862398595, 0.190863123396412, 0.462323304964229, 0.501146966125816, 0.520007181214169, 0.564923198428005, 0.649097554385662, 0.567103909794241, 0.803667107596993, 0.590206037275493, 0.689842557767406, 0.412291133077815, 0.509641209617257, 0.0209674017969519, 0.88397897221148, 0.765795998740941, 0.252576215192676, 0.651508895447478, 0.306829280918464, 0.185269475681707, 0.71143921231851, 0.548262833384797, 0.352060183184221, 0.472812375053763, 0.275556553853676, 0.734964651986957, 0.523070408729836, 0.742901084013283, 0.745984371984378, 0.025135435629636, 0.635091364383698, 0.555273689562455, 0.60693351412192, 0.602010689210147, 0.131266884040087, 0.104975792113692, 0.326069595525041, 0.740822408115491, 0.198113218182698, 0.77243953733705, 0.847749983659014, 0.819884012220427, 0.508232545806095, 0.397203155560419, 0.199203212046996, 0.797252381918952, 0.307308677351102, 0.146779833827168, 0.329096749192104, 0.0053351204842329, 0.119237521430477, 0.0701981610618532, 0.0449260456953198, 0.502128139138222, 0.682406621985137, 0.949905222048983, 0.990762530360371, 0.584857220062986, 0.665898126550019, 0.920643171295524, 0.0649762516841292, 0.0567484009079635, 0.34496046230197, 0.293499368242919, 0.733789156191051, 0.97696074959822, 0.186083873966709, 0.661094717914239, 0.346165475901216, 0.225813308497891, 0.838807633612305, 0.547229760326445, 0.693947496823967, 0.392685191705823, 0.384549561189488, 0.796505460049957, 0.581522558117285, 0.145260240184143, 0.750296842074022, 0.29040913679637, 0.65269380598329, 0.717949920101091, 0.184230447979644, 0.351494160713628, 0.650116921868175, 0.335423482116312, 0.899857638869435, 0.21809272095561, 0.936820335919037, 0.695472309598699, 0.938670792384073, 0.639230779139325, 0.0348742720670998, 0.586689950199798, 0.585411561885849, 0.260398308280855, 0.844541017664596, 0.541734155965969, 0.764398477505893, 0.896406861254945, 0.686280905967578, 0.151093304157257, 0.67225520638749, 0.471414312254637, 0.0663757910951972, 0.835177519125864, 0.111039395211264, 0.956101143965498, 0.642315058270469, 0.744551684940234, 0.202209664741531, 0.852585457498208, 0.590050214901567, 0.257464315509424, 0.0485539957880974, 0.615514848381281, 0.811861861031502, 0.524702502880245, 0.984270685119554, 0.499994627432898, 0.875676695257425, 0.305305047892034, 0.691856879042462, 0.64961622026749, 0.947139994474128, 0.171197948046029, 0.0824002479203045, 0.859867240069434, 0.228181310463697, 0.566435108659789, 0.973865487845615, 0.724375836318359, 0.940533125540242, 0.275708921486512, 0.57892808620818, 0.265651260735467, 0.9981757204514, 0.802141561638564, 0.889236446470022, 0.215859720017761, 0.214633448515087, 0.998030096758157, 0.684333952376619, 0.0086820344440639, 0.440940480679274, 0.99374751909636, 0.591970882145688, 0.678340346319601, 0.0301579283550382, 0.534216424450278, 0.809104614891112, 0.94899785425514, 0.451514917891473, 0.229145936667919, 0.0857485535088927, 0.771948733832687, 0.00728950090706348, 0.547074251575395, 0.610671214526519, 0.483363758539781, 0.977494483347982, 0.830507989274338, 0.456219208892435, 0.604671768378466, 0.911703945370391, 0.00769801833666861, 0.0848501976579428, 0.0201672618277371, 0.783988921903074, 0.356370847672224, 0.0313490005210042, 0.772229343419895, 0.78000157745555, 0.387067530537024, 0.223980102455243, 0.749804693041369, 0.857148488517851, 0.115232218522578, 0.671802728669718, 0.468645263696089, 0.363730967510492, 0.228265785612166, 0.816435700515285, 0.410148762864992, 0.68911273824051, 0.423632357036695, 0.385312004946172, 0.185305827762932, 0.347943822387606, 0.26281138905324, 0.681595667032525, 0.838560755364597, 0.614662179956213, 0.177759745391086, 0.233822337351739, 0.160610055550933, 0.193676067050546, 0.00340610509738326, 0.281589049845934, 0.0092222869861871, 0.575361076975241, 0.337534205522388, 0.209672437282279, 0.475082165794447, 0.341732465894893, 0.13111048610881, 0.419216299196705, 0.125647888984531, 0.0257873123046011, 0.417283693095669, 0.0636545692104846, 0.866621513850987, 0.593349575065076, 0.914505168562755, 0.968387644505128, 0.05062130629085, 0.28176791081205, 0.105222090613097, 0.134627741295844, 0.192412375938147, 0.760852600913495, 0.238934814929962, 0.147984366165474, 0.541062023956329, 0.84454024140723, 0.799618279095739, 0.780540036968887, 0.363787683425471, 0.202850539237261, 0.669798091519624, 0.886520284926519, 0.445746072335169, 0.0205153631977737, 0.182691988768056, 0.928953429684043, 0.00157543644309044, 0.727433199761435, 0.530162011738867, 0.593859041342512, 0.751685525057837, 0.669151596026495, 0.793493767268956, 0.592926530633122, 0.0089231280144304, 0.538555353414267, 0.550756573211402, 0.0284826112911105, 0.436842594761401, 0.436036470811814, 0.535622813506052, 0.280515288934112, 0.701827713288367, 0.0603534092660993, 0.571080398280174, 0.937864253064618, 0.941148055484518, 0.0736383756157011, 0.189283938612789, 0.278909419896081, 0.997881927993149, 0.625220316462219, 0.602562202839181, 0.85343104088679, 0.323151511140168, 0.692986620124429, 0.0888787612784654, 0.284639050951228, 0.8474933616817, 0.0955237015150487, 0.910568720893934, 0.657226078445092, 0.264876997331157, 0.148349090712145, 0.503495522309095, 0.273692655842751, 0.824444289086387, 0.174309683265164, 0.10006021335721, 0.263755331048742, 0.710263483691961, 0.369161438196898, 0.846208694390953, 0.422358606243506, 0.613178660394624, 0.442004077835009, 0.909304512664676, 0.917323393747211, 0.853543300414458, 0.210952980211005, 0.359097968554124, 0.157947130966932, 0.448106565978378, 0.321996413636953, 0.669583174632862, 0.511823168257251, 0.955190336098894, 0.535673343576491, 0.171886418946087, 0.778302463935688, 0.489757213043049, 0.0223054855596274, 0.850990085629746, 0.447587976465002, 0.284467703197151, 0.659318662714213, 0.886985338525847, 0.2705631935969, 0.415176111971959, 0.436109684640542, 0.790490825893357, 0.635096330894157, 0.599321972345933, 0.858126113656908, 0.927643255330622, 0.399829821893945, 0.481699954485521, 0.983611203497276, 0.477281948318705, 0.9006813892629, 0.715750393457711, 0.987675331300125, 0.612305964808911, 0.325083426898345, 0.737518916837871, 0.843966565094888, 0.0674192304722965, 0.620740816695616, 0.72668613283895, 0.134120569098741, 0.920677061658353, 0.326856863219291, 0.243085787864402, 0.869809092720971, 0.65096335997805, 0.680623615859076, 0.83814689097926, 0.517129164654762, 0.22469749301672, 0.131095506018028, 0.821202026680112, 0.131159810349345, 0.607636734377593, 0.713033704319969, 0.479691217653453, 0.864707166561857, 0.841359662823379, 0.170258817262948, 0.347030242206529, 0.620412183459848, 0.963180393213406, 0.0368542631622404, 0.751684188609943, 0.821201306534931, 0.506466376595199, 0.357957704458386, 0.0551878632977605, 0.804044840857387, 0.496529700467363, 0.337760641472414, 0.764270386425778, 0.90140912681818, 0.479193103499711, 0.344308067113161, 0.489079599268734, 0.453782195225358, 0.154929247219115, 0.497734451899305, 0.830463097197935, 0.489010628312826, 0.18102588923648, 0.166730530560017, 0.336230527376756, 0.719154513673857, 0.411083461018279, 0.0870308505836874, 0.371234548976645, 0.985779159935191, 0.786128085572273, 0.244802305940539, 0.282681001583114, 0.352139938622713, 0.730121547356248, 0.279444871935993, 0.304560498334467, 0.565175200812519, 0.794138974975795, 0.182650593109429, 0.199398258933797, 0.59422831539996, 0.872430627699941, 0.345254213782027, 0.985628178110346, 0.755010990658775, 0.9995037028566, 0.109075453830883, 0.0827170386910439, 0.944558904971927, 0.316450600977987, 0.613627026323229, 0.324247510870919, 0.60964174522087, 0.4357713337522, 0.640956471441314, 0.622685300186276, 0.271656002150849, 0.321775593794882, 0.791006790241227, 0.760036299005151, 0.0125722368247807, 0.418849518988281, 0.732083063572645, 0.127892419695854, 0.811107192654163, 0.436911109136418, 0.878763943212107, 0.210802522720769, 0.428816736210138, 0.263784283073619, 0.55096147605218, 0.00658412766642869, 0.878501063911244, 0.544453230919316, 0.415108400629833, 0.356027404079214, 0.128325364319608, 0.584777529584244, 0.329645261634141, 0.491703604115173, 0.993227555416524, 0.98839941737242, 0.990954062901437, 0.495830727275461, 0.271479320945218, 0.962515514343977, 0.797615485265851, 0.660060367546976, 0.863920476520434, 0.0676849773153663, 0.886744817020372, 0.0869222974870354, 0.777475133072585, 0.426183409523219, 0.017279535997659, 0.93702796055004, 0.407067511929199, 0.818093952955678, 0.174666524631903, 0.0212195052299649, 0.355476544471458, 0.127436563372612, 0.492130193160847, 0.598317288095132, 0.0843912751879543, 0.0436145768035203, 0.621073262533173, 0.293654473964125, 0.690387488342822, 0.492559402715415, 0.417443345300853, 0.0920450289268047, 0.334763678489253, 0.90665902197361, 0.850888444809243, 0.800898156361654, 0.5673239659518, 0.382200802443549, 0.0193239971995354, 0.246499913046136, 0.288420708850026, 0.470918375998735, 0.871305386768654, 0.811009631957859, 0.45502445125021, 0.816309611313045, 0.515951848588884, 0.602965562837198, 0.804503378691152, 0.956501983338967, 0.566461805254221, 0.833880389342085, 0.245733676711097, 0.114893596386537, 0.487205882556736, 0.871003806823865, 0.0310964758973569, 0.468159379670396, 0.297316632466391, 0.153489704243839, 0.807778357760981, 0.185858204262331, 0.783149952301756, 0.0769749032333493, 0.438188479980454, 0.220023768953979, 0.0916111960541457, 0.40475213364698, 0.460640790406615, 0.209776928182691, 0.673804820049554, 0.160605198005214, 0.877638026373461, 0.320863935863599, 0.622471495065838, 0.757471777731553, 0.91328929620795, 0.929171101655811, 0.374533589463681, 0.262555107008666, 0.936005780706182, 0.98309379001148, 0.489040536805987, 0.224842398893088, 0.262857019202784, 0.949720339616761, 0.315314318286255, 0.82716026599519, 0.830440465593711, 0.664254550123587, 0.603915638988838, 0.114426568383351, 0.0892527953255922, 0.891117219114676, 0.679889178834856, 0.697479327674955, 0.591649477137253, 0.376564126694575, 0.667763240635395, 0.175413054181263, 0.327615150250494, 0.876706693554297, 0.961855879053473, 0.0523357135243714, 0.524481996661052, 0.357952268794179, 0.114158331649378, 0.49246078543365, 0.12605739175342, 0.331631509121507, 0.210683841025457, 0.410896329209208, 0.7278327930253, 0.0659876961726695, 0.443817230174318, 0.0105216549709439, 0.146687521366403, 0.0721488478593528, 0.812878718134016, 0.360742731252685, 0.702489458955824, 0.540977456141263, 0.304459790699184, 0.12065661791712, 0.140375901712105, 0.885186838684604, 0.39649873948656, 0.33665336528793, 0.568767318269238, 0.227991473395377, 0.351324187824503, 0.5942509889137, 0.574141363147646, 0.609925560420379, 0.979146542726085, 0.713803424499929, 0.985914314864203, 0.342103280359879, 0.0493026291951537, 0.670093417400494, 0.614890307420865, 0.757787862792611, 0.354943498736247, 0.470823898212984, 0.190468366490677, 0.180673186667264, 0.258185529382899, 0.144511143909767, 0.216533446218818, 0.0392066079657525, 0.608593434561044, 0.746887418674305, 0.908565161284059, 0.138278678990901, 0.744614053284749, 0.877130803652108, 0.627210536506027, 0.85312277963385, 0.248976753558964, 0.351620918838307, 0.802123087923974, 0.618778681149706, 0.979238719213754, 0.527631651610136, 0.314052432775497, 0.910631059668958, 0.761038681957871, 0.105642720591277, 0.262125560315326, 0.929216892691329, 0.360194638604298, 0.272630422376096, 0.758303255774081, 0.120735815260559, 0.228261379059404, 0.0646116565912962, 0.6915870369412, 0.795619363198057, 0.83071054937318, 0.432599925203249, 0.759454179322347, 0.799309197114781, 0.469810077920556, 0.302143429871649, 0.648127724649385, 0.0581815482582897, 0.755429731449112, 0.826911469222978, 0.391930246492848, 0.0699666009750217, 0.272376773646101, 0.760664598084986, 0.130423314170912, 0.431030840147287, 0.496877519180998, 0.821703399298713, 0.0100158331915736, 0.332261150470003, 0.898630318930373, 0.83180737728253, 0.0693970918655396, 0.364471346372738, 0.974293649429455, 0.852367438608781, 0.604516058694571, 0.573901277966797, 0.188153061084449, 0.461867716861889, 0.234223118750378, 0.114163143793121, 0.517937210155651, 0.957693409407511, 0.828540002461523, 0.0180606378708035, 0.155524155357853, 0.986596627160907, 0.627513964194804, 0.754470132291317, 0.0816371464170516, 0.359466931549832, 0.85710916807875, 0.100963250268251, 0.343091274145991, 0.630265658022836, 0.187501690583304, 0.406499820761383, 0.968732046196237, 0.768846731167287, 0.662728684721515, 0.945256893523037, 0.0595155721530318, 0.202088307123631, 0.757733877049759, 0.708705253899097, 0.452036932110786, 0.338992069242522, 0.932918353937566, 0.861345540499315, 0.664538627956063, 0.924660959746689, 0.512407177360728, 0.227494287537411, 0.884186332579702, 0.453545617172495, 0.803610211005434, 0.561296524479985, 0.90043152263388, 0.841235359665006, 0.0218471572734416, 0.199693278642371, 0.445917607517913, 0.168094933731481, 0.334469676716253, 0.234457859070972, 0.408613098552451, 0.473678894573823, 0.545817021513358, 0.528108806814998, 0.555053427582607, 0.41802521282807, 0.755254870280623, 0.168757189530879, 0.00308841303922236, 0.552166923880577, 0.590335902990773, 0.803637700155377, 0.621779754059389, 0.281861263327301, 0.388060290832072, 0.613821845036, 0.233192824292928, 0.00146270240657032, 0.531414496246725, 0.301926037063822, 0.950179205974564, 0.0705745436716825, 0.68007143237628, 0.952487320173532, 0.298776651965454, 0.50964452768676, 0.343977516051382, 0.875302866334096, 0.625747997546569, 0.420583335217088, 0.785828031832352, 0.94993125111796, 0.298471388407052, 0.220887282863259, 0.311348300660029, 0.695207278942689, 0.410034971078858, 0.848853385308757, 0.982433591969311, 0.823073085630313, 0.708653011824936, 0.441271203570068, 0.711133639328182, 0.187097328016534, 0.87804185715504, 0.100282375933602, 0.193222832633182, 0.110730886925012, 0.930115651339293, 0.35559572535567, 0.390074053080752, 0.568954457063228, 0.678489098092541, 0.574205453740433, 0.495365530485287, 0.83155357488431, 0.25580694922246, 0.28870277130045, 0.905404130462557, 0.602930790511891, 0.964311280520633, 0.14968887809664, 0.0345470882020891, 0.298657424980775, 0.381702641956508, 0.519340321887285, 0.923516572220251, 0.108191382372752, 0.586938820080832, 0.0640069998335093, 0.535426516085863, 0.725104424869642, 0.824675458483398, 0.0597329833544791, 0.403060883982107, 0.826684100087732, 0.433433753671125, 0.654455939307809, 0.236722751287743, 0.626350827515125, 0.33179133804515, 0.878774685785174, 0.478058483451605, 0.0283857004251331, 0.184247707482427, 0.496152664301917, 0.185747111449018, 0.226192387519404, 0.178591825766489, 0.00096238637343049, 0.611378089524806, 0.23054240969941, 0.586107048438862, 0.128616735106334, 0.0487355305813253, 0.527958194958046, 0.679715928854421, 0.758881898131222, 0.3308006352745, 0.835491568315774, 0.509642125340179, 0.55941771180369, 0.802393204066902, 0.241020564455539, 0.665586761198938, 0.331096068723127, 0.281442808685824, 0.920128356665373, 0.0242515790741891, 0.114211968146265, 0.0559281199239194, 0.456008055945858, 0.0533575734589249, 0.348146997159347, 0.295905384467915, 0.637316369684413, 0.142845072550699, 0.036300852894783, 0.708561948267743, 0.172357433708385, 0.124244898324832, 0.296556779416278, 0.316187723074108, 0.919586347183213, 0.831615468254313, 0.306839636527002, 0.13727489579469, 0.872984929941595, 0.427575533743948, 0.276652917498723, 0.389649858232588, 0.0434066478628665, 0.946206420427188, 0.203245186479762, 0.111963690491393, 0.530163813848048, 0.647847692715004, 0.963890448212624, 0.428996141767129, 0.876209651818499, 0.305686295265332, 0.497765189269558, 0.647584401303902, 0.834353565471247, 0.921636421000585, 0.0267052904237062, 0.0285582225769758, 0.845971006900072, 0.625793483341113, 0.0421565526630729, 0.672119665658101, 0.968265606556088, 0.404815388144925, 0.0026925343554467, 0.789196252124384, 0.102081243414432, 0.912044989643618, 0.98252965672873, 0.883580181747675, 0.153021680656821, 0.438747941050678, 0.0578767901752144, 0.42605418455787, 0.45082975178957, 0.828206003177911, 0.089122753823176, 0.88586020283401, 0.737388913985342, 0.299373862333596, 0.994218221865594, 0.950738520128652, 0.957855986664072, 0.844872489338741, 0.873361982870847, 0.664381322683766, 0.0897392614278942, 0.0665703311096877, 0.364557095104828, 0.484666807111353, 0.588397777173668, 0.636721211718395, 0.707946713780984, 0.968993826303631, 0.606920960359275, 0.500432366970927, 0.227967844111845, 0.155387134756893, 0.12902299920097), actual = c(0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"

# get a combination of cutpoints from both user-defined functions and inbuilt methods
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
  nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
)
#> Error in fx_roc01(predicted = c(0.95392177044414, 0.448208634508774, 0.0270868674851954, 0.312464930349961, 0.374010580359027, 0.762560096569359, 0.994020394282416, 0.931702611269429, 0.110634732525796, 0.723226918606088, 0.677323925076053, 0.464009287999943, 0.760693490505219, 0.572973643662408, 0.150719033088535, 0.782009037677199, 0.689068002859131, 0.973873599199578, 0.264332292834297, 0.827052976470441, 0.275190070969984, 0.892771044746041, 0.743128176080063, 0.389291333733127, 0.257925776997581, 0.8720326453913, 0.54515916807577, 0.449880579253659, 0.230599802453071, 0.0469215563498437, 0.193160629132763, 0.714908554218709, 0.417437852127478, 0.554366349475458, 0.697381522972137, 0.490321051795036, 0.724393639015034, 0.530261610168964, 0.242871329886839, 0.082656008657068, 0.619269102578983, 0.368588034296408, 0.445287212030962, 0.244780165608972, 0.469249901827425, 0.519291918491945, 0.872115530539304, 0.914415181847289, 0.576115989591926, 0.305186428828165, 0.967851023888215, 0.0462249761912972, 0.895993108395487, 0.638897842494771, 0.600174552295357, 0.796648465096951, 0.291167046874762, 0.85140435397625, 0.727506811264902, 0.98883752967231, 0.975900620920584, 0.587031626608223, 0.943989926017821, 0.97247190377675, 0.331973553169519, 0.118335549486801, 0.515561492880806, 0.722935039550066, 0.542774250730872, 0.932126874104142, 0.489039185456932, 0.573097415035591, 0.814293791539967, 0.673834793036804, 0.931312846951187, 0.224417308811098, 0.907946768449619, 0.0473488674033433, 0.94052480859682, 0.47075440059416, 0.188107250258327, 0.352237876970321, 0.633567616343498, 0.564717964269221, 0.621177914086729, 0.183029039064422, 0.22397765214555, 0.617047521751374, 0.36407410656102, 0.848919195123017, 0.821185347624123, 0.94977107946761, 0.96609840542078, 0.738501994404942, 0.804637480992824, 0.324505825759843, 0.699429291300476, 0.742739485809579, 0.225012851180509, 0.880483536049724, 0.0283928043209016, 0.931644033873454, 0.573924665804952, 0.425096873892471, 0.873483943287283, 0.504262586124241, 0.707543053897098, 0.731148096267134, 0.695272200042382, 0.565705072134733, 0.839416237315163, 0.854858984705061, 0.356831853743643, 0.3583366968669, 0.0124742113985121, 0.562798132188618, 0.578789162682369, 0.704509716713801, 0.648126327898353, 0.690166405867785, 0.363105172757059, 0.756288834614679, 0.21818868117407, 0.471511785173789, 0.830583182163537, 0.0816900080535561, 0.808214358054101, 0.238532107323408, 0.988307715160772, 0.292504418641329, 0.413411591202021, 0.736935928929597, 0.414706182898954, 0.666836815886199, 0.224415857344866, 0.387034191284329, 0.870875645661727, 0.555063757114112, 0.847957621328533, 0.547497714869678, 0.296962636522949, 0.316252124728635, 0.29712104704231, 0.0851685288362205, 0.305096935480833, 0.812476988416165, 0.115293657174334, 0.595611366676167, 0.074240337125957, 0.0831423727795482, 0.387614593375474, 0.555880051339045, 0.463226814521477, 0.5913009212818, 0.760815987596288, 0.993863243609667, 0.882550185080618, 0.0402875868603587, 0.569576953537762, 0.0509860219899565, 0.978969879914075, 0.636136224260554, 0.398667011409998, 0.0497206572908908, 0.947778065456077, 0.986051214626059, 0.479523075278848, 0.476336157415062, 0.875680567231029, 0.0398007486946881, 0.338993408717215, 0.386998119298369, 0.88375075859949, 0.0666755959391594, 0.310872238595039, 0.669798787683249, 0.493629577336833, 0.53297019796446, 0.841235504951328, 0.84633452934213, 0.603294087108225, 0.766719637671486, 0.562451550504193, 0.120406513102353, 0.704185891430825, 0.112970114918426, 0.199042586842552, 0.708155807107687, 0.9178075583186, 0.947394807590172, 0.224459138698876, 0.996059477096424, 0.274510144256055, 0.554936560802162, 0.893440310843289, 0.242682221112773, 0.869353267131373, 0.552072071935982, 0.337965423008427, 0.753823082428426, 0.380773828132078, 0.225850170478225, 0.0631416407413781, 0.035363185685128, 0.343772424152121, 0.850413308478892, 0.324566725408658, 0.758387070382014, 0.515551632270217, 0.0289400399196893, 0.766118261264637, 0.480022622505203, 0.286896881647408, 0.476576584624127, 0.578515741974115, 0.298033373663202, 0.342470785370097, 0.379823049996048, 0.0926486370153725, 0.708783565787598, 0.405673941131681, 0.804931717691943, 0.236376116750762, 0.663599726045504, 0.037486799294129, 0.825246879365295, 0.0331633093301207, 0.394040404353291, 0.470865379320458, 0.581663032993674, 0.997080007568002, 0.021587701048702, 0.778782083187252, 0.894903084961697, 0.62262161122635, 0.491981435567141, 0.483395426999778, 0.759226365480572, 0.901995153632015, 0.338175330776721, 0.718513572588563, 0.283959623426199, 0.986670962069184, 0.868605944560841, 0.949325458845124, 0.660575847374275, 0.653229519957677, 0.157853540265933, 0.0957874974701554, 0.884381188778207, 0.098823441658169, 0.164384096860886, 0.425716221099719, 0.678098261123523, 0.881315565668046, 0.909516155952588, 0.595481478841975, 0.599045276874676, 0.354431133018807, 0.536809108685702, 0.0204283259809017, 0.626410849159583, 0.0994539232924581, 0.555199061753228, 0.757537386845797, 0.93317555822432, 0.990495617035776, 0.68116157525219, 0.286483302945271, 0.913638089550659, 0.554285926977172, 0.867340955184773, 0.572826200164855, 0.769890369148925, 0.745343824615702, 0.062036223243922, 0.602984179044142, 0.0832468441221863, 0.548292555147782, 0.787109596189111, 0.247856200207025, 0.236723284004256, 0.673211005749181, 0.871645119041204, 0.383594762999564, 0.236125354887918, 0.26479925843887, 0.639583244221285, 0.898541283560917, 0.683867615880445, 0.659976146183908, 0.201700175181031, 0.17970937397331, 0.284552552737296, 0.440125150140375, 0.989611474564299, 0.952993207145482, 0.951802115887403, 0.292846935801208, 0.0791806364431977, 0.609316575108096, 0.870226508006454, 0.0638812573160976, 0.23843178874813, 0.690071495482698, 0.24151929654181, 0.41513179638423, 0.502102697035298, 0.980931859230623, 0.187816001940519, 0.85812796629034, 0.84606016613543, 0.175825901096687, 0.489260280970484, 0.0856519450899214, 0.15342704532668, 0.307175467023626, 0.450554560637102, 0.0954182576388121, 0.704218784114346, 0.518729069037363, 0.717157990438864, 0.489642962347716, 0.555392521899194, 0.447089770110324, 0.486188290640712, 0.182296962477267, 0.553956833435223, 0.458408696576953, 0.855392566649243, 0.763228131690994, 0.456659199437127, 0.259058269904926, 0.992431030143052, 0.875758427660912, 0.057351116091013, 0.268993757897988, 0.149863939499483, 0.917890511453152, 0.512903506169096, 0.240816539619118, 0.501868763938546, 0.899600558681414, 0.817811503773555, 0.829001636942849, 0.0141677737701684, 0.352093641646206, 0.740824914071709, 0.408628272823989, 0.826643102569506, 0.809073497774079, 0.111898281145841, 0.178306270157918, 0.631438093259931, 0.117671165848151, 0.919767881976441, 0.993820427684113, 0.846939872950315, 0.656121122417971, 0.75516741652973, 0.921485370257869, 0.802752825431526, 0.0184027764480561, 0.605465486645699, 0.561869250843301, 0.332315589534119, 0.300114005571231, 0.412497795419767, 0.125018023420125, 0.990632067201659, 0.00277790054678917, 0.227618644712493, 0.505978730507195, 0.877471745014191, 0.57197172078304, 0.575703122653067, 0.368136106757447, 0.166029269807041, 0.30691988626495, 0.805915336823091, 0.4447358364705, 0.603211194975302, 0.0100145500618964, 0.353463196195662, 0.394041391322389, 0.106245225295424, 0.581637792987749, 0.0209359570872039, 0.762188960798085, 0.10815327335149, 0.761099917581305, 0.842418174259365, 0.787208508467302, 0.171650050207973, 0.575807052198797, 0.906851674197242, 0.633713058661669, 0.107691245619208, 0.864335746504366, 0.487701000645757, 0.719550532056019, 0.761155858635902, 0.259229828137904, 0.306443256093189, 0.981644687242806, 0.476794175338, 0.503503609914333, 0.674510363256559, 0.881803320720792, 0.402258340734988, 0.18936653342098, 0.00719878450036049, 0.741551893763244, 0.336345520801842, 0.734765686560422, 0.358470146078616, 0.529710022499785, 0.7644828397315, 0.433423662325367, 0.681007634382695, 0.348257374484092, 0.618165794759989, 0.930115365190431, 0.156686106696725, 0.689936604816467, 0.904002557741478, 0.982049233512953, 0.532924651633948, 0.324978902004659, 0.549126014113426, 0.882867860607803, 0.801305397879332, 0.640604358399287, 0.0148338556755334, 0.719052298460156, 0.138138834852725, 0.370501160155982, 0.98390534799546, 0.985266339499503, 0.28880266030319, 0.0796563306357712, 0.0696727754548192, 0.808540128869936, 0.573813927359879, 0.918623236240819, 0.40908720321022, 0.958784570917487, 0.948390351375565, 0.714043829822913, 0.233863855013624, 0.746340417535976, 0.949958555866033, 0.973201578715816, 0.509500867221504, 0.590598380193114, 0.626137979095802, 0.956263393396512, 0.329663357464597, 0.963153187185526, 0.77027030964382, 0.815819204086438, 0.832929047057405, 0.0387769781518728, 0.992728653363883, 0.138488719472662, 0.983745038509369, 0.645176357822493, 0.890656743198633, 0.0768370819278061, 0.459420280763879, 0.556014743633568, 0.254792039515451, 0.35169401136227, 0.528484009206295, 0.406059542903677, 0.389964426867664, 0.377409445354715, 0.684875444276258, 0.937554783420637, 0.125583315035328, 0.724634517449886, 0.408404387533665, 0.302393236197531, 0.390497365966439, 0.486830823821947, 0.850951658096164, 0.065549022750929, 0.417581103742123, 0.835076132556424, 0.559288707561791, 0.373631848720834, 0.104444408556446, 0.796680721454322, 0.840642758877948, 0.0485098555218428, 0.71074540540576, 0.0665934230200946, 0.452565061161295, 0.319251748267561, 0.776146202348173, 0.41894514625892, 0.748611016897485, 0.301067522494122, 0.861751779913902, 0.83867881144397, 0.832865367177874, 0.498801694717258, 0.248949346132576, 0.625498958863318, 0.805195474997163, 0.840655814856291, 0.47377585968934, 0.0452975444495678, 0.536885789595544, 0.448440016247332, 0.875013964716345, 0.707804388832301, 0.707231404725462, 0.290162386372685, 0.97164366627112, 0.497302305418998, 0.363340763142332, 0.632141254842281, 0.657176745822653, 0.5928277191706, 0.932919291546568, 0.991049832198769, 0.747818121453747, 0.902658168924972, 0.27350633777678, 0.124323650496081, 0.337975647998974, 0.595657059457153, 0.622849642764777, 0.490782752167434, 0.881586844567209, 0.117060035467148, 0.638999679358676, 0.657827848568559, 0.564410293474793, 0.383561118505895, 0.383091382449493, 0.111184020293877, 0.96049310429953, 0.0555843852926046, 0.999079468194395, 0.221924037439749, 0.606887325644493, 0.446580953663215, 0.160909163765609, 0.107782390434295, 0.365347869694233, 0.959607374854386, 0.111779827624559, 0.230004532728344, 0.184092711657286, 0.35261666146107, 0.622050385922194, 0.844117138069123, 0.23121522529982, 0.0918678662274033, 0.292846794473007, 0.403359015006572, 0.0212086625397205, 0.710785044590011, 0.274726219242439, 0.0220727897249162, 0.190852003637701, 0.785697978688404, 0.0287587761413306, 0.225082878256217, 0.153478813124821, 0.218588055577129, 0.931217759847641, 0.0984466595109552, 0.160434607882053, 0.843959147110581, 0.0539984283968806, 0.0500226262956858, 0.977120511000976, 0.886250207433477, 0.932292006677017, 0.105596865527332, 0.9434682095889, 0.358370498288423, 0.128870363347232, 0.82645116536878, 0.441675641341135, 0.416865685023367, 0.612892851000652, 0.554529616143554, 0.0137557715643197, 0.33788674371317, 0.236288583371788, 0.88938637310639, 0.91177484113723, 0.857980493921787, 0.242132487241179, 0.823638271773234, 0.444390260381624, 0.2005197359249, 0.43172637722455, 0.85957630421035, 0.627201276831329, 0.707803406752646, 0.571518149459735, 0.655664514284581, 0.392644048202783, 0.179659378482029, 0.480551759479567, 0.668806851143017, 0.00694709876552224, 0.997687928611413, 0.0688496548682451, 0.169651831733063, 0.340740051586181, 0.714877606369555, 0.229822572786361, 0.222771419677883, 0.933238807599992, 0.873943994054571, 0.519540133420378, 0.100505406735465, 0.544299082830548, 0.720182369928807, 0.6897099877242, 0.710416536545381, 0.661384833976626, 0.835523371351883, 0.559458846691996, 0.735170545056462, 0.645231382921338, 0.517036242876202, 0.606005186680704, 0.381827254779637, 0.582628410309553, 0.739997710799798, 0.411296576494351, 0.205508053535596, 0.0233304719440639, 0.154921570094302, 0.568840727908537, 0.629065089160576, 0.228474416071549, 0.99278173292987, 0.0122038063127548, 0.645099396584556, 0.982464784523472, 0.75593027472496, 0.385795035166666, 0.0882505599875003, 0.619783494155854, 0.594573868671432, 0.671272209146991, 0.788127640960738, 0.449218205176294, 0.125611432828009, 0.181707219220698, 0.156297316309065, 0.069544873200357, 0.78730232338421, 0.127441613702103, 0.587336444295943, 0.153287045657635, 0.520016293972731, 0.518447174225003, 0.810656490037218, 0.399216897087172, 0.583093363791704, 0.450175904436037, 0.893696022685617, 0.516466134926304, 0.35275375796482, 0.616691372357309, 0.160249023232609, 0.900125857209787, 0.211095432052389, 0.451487764017656, 0.452967888908461, 0.621726003009826, 0.0860377156641334, 0.320382753154263, 0.991191994166002, 0.177653705468401, 0.0395130345132202, 0.500591187272221, 0.850309741450474, 0.481186116812751, 0.826985011342913, 0.553660382749513, 0.404830258106813, 0.367257896577939, 0.128533743787557, 0.725018117809668, 0.929977179504931, 0.221120818052441, 0.166530205169693, 0.541213060030714, 0.677352456608787, 0.355230290442705, 0.101814652327448, 0.0621043336577713, 0.483566273702309, 0.332433048635721, 0.144014274002984, 0.242219908861443, 0.390183951472864, 0.00883543654344976, 0.445334241725504, 0.370063247857615, 0.50120793050155, 0.846330758417025, 0.382669016253203, 0.322618347592652, 0.382451640442014, 0.0487738088704646, 0.000559384468942881, 0.59502287581563, 0.693871656432748, 0.594431041739881, 0.379488051170483, 0.153814671561122, 0.54212937830016, 0.591316811507568, 0.838857745751739, 0.858507097465917, 0.701488207792863, 0.845201757038012, 0.389302667230368, 0.377124967752025, 0.137570964870974, 0.953620044514537, 0.518870074301958, 0.68237853422761, 0.38216333813034, 0.451726326020434, 0.624546718550846, 0.854567285394296, 0.0267164036631584, 0.200257339514792, 0.183637893293053, 0.769589643925428, 0.8977469811216, 0.963413840159774, 0.721836498240009, 0.883468134794384, 0.222086797701195, 0.662221425911412, 0.053293242584914, 0.481304476503283, 0.584288695361465, 0.483105749124661, 0.789375885156915, 0.700753993354738, 0.190443291794509, 0.937207114417106, 0.304137283470482, 0.632712226593867, 0.708174541359767, 0.782063946593553, 0.767743366770446, 0.00235058460384607, 0.282703186385334, 0.354706982150674, 0.247077039675787, 0.071677393047139, 0.998484137933701, 0.72422236087732, 0.499825933715329, 0.962556596612558, 0.681642477866262, 0.521872260374948, 0.930221057729796, 0.849517413880676, 0.699770094361156, 0.533122876426205, 0.54161281324923, 0.394316250225529, 0.0700038145296276, 0.406233927002177, 0.982967500342056, 0.511063785292208, 0.421083862427622, 0.943471364211291, 0.702753707533702, 0.212342031067237, 0.27984134433791, 0.375090099172667, 0.884403531905264, 0.387930921046063, 0.476140171522275, 0.692606335505843, 0.211380551103503, 0.836063615744933, 0.563172003487125, 0.566549305105582, 0.184976289281622, 0.623277766397223, 0.204355993075296, 0.56962132640183, 0.969529064837843, 0.845789787126705, 0.533509102882817, 0.374031426152214, 0.73041637102142, 0.864077596459538, 0.803917703451589, 0.37312152213417, 0.416990536265075, 0.917091441340744, 0.714573037577793, 0.937610322609544, 0.281617351807654, 0.555251316865906, 0.16548340395093, 0.478522335644811, 0.37876583612524, 0.0520021233242005, 0.956744392868131, 0.0803927790839225, 0.993022685870528, 0.748503470094875, 0.234840938821435, 0.459694660734385, 0.0353945791721344, 0.251753443619236, 0.721953386440873, 0.00822188798338175, 0.803747632307932, 0.977728196885437, 0.654888457851484, 0.102512364275753, 0.338691339828074, 0.0389515543356538, 0.665638877544552, 0.792956927558407, 0.269959869561717, 0.978103853296489, 0.232835488626733, 0.427874536952004, 0.270039998460561, 0.298475892050192, 0.188851399580017, 0.282581120962277, 0.307444769190624, 0.685975810745731, 0.146726256702095, 0.60177163197659, 0.571354366838932, 0.919172737747431, 0.0104328731540591, 0.284845159389079, 0.628395191160962, 0.704525556880981, 0.126142370747402, 0.868731233291328, 0.242769131902605, 0.294888264965266, 0.892542588058859, 0.617673375178128, 0.326807370176539, 0.526563023682684, 0.433942462783307, 0.839977380353957, 0.0492059183306992, 0.991210207110271, 0.0634351919870824, 0.396323537454009, 0.683265100000426, 0.405670127598569, 0.891173214418814, 0.230037525063381, 0.937082352582365, 0.0629696603864431, 0.783495836192742, 0.437238272977993, 0.134081751806661, 0.459364474285394, 0.68682372639887, 0.502273478545249, 0.797684699296951, 0.622882501920685, 0.981579658342525, 0.68307189992629, 0.74890649667941, 0.326838234439492, 0.029294399311766, 0.592875162605196, 0.997856359463185, 0.919762917095795, 0.512601620750502, 0.933649256126955, 0.110269572352991, 0.859257325297222, 0.632527868496254, 0.82490489911288, 0.427866259356961, 0.781971692340448, 0.437545914202929, 0.926329316105694, 0.666486889356747, 0.0954146075528115, 0.532410041661933, 0.866993244970217, 0.0989461098797619, 0.675321348244324, 0.760120302671567, 0.398194221081212, 0.281501155113801, 0.741546059027314, 0.149954436346889, 0.0480502068530768, 0.42382941651158, 0.434847516240552, 0.933499967446551, 0.459481004858389, 0.922503989189863, 0.303984656231478, 0.312268117908388, 0.577410660684109, 0.199493832420558, 0.527752980357036, 0.619486727518961, 0.817761204438284, 0.00117613677866757, 0.739626267692074, 0.794578975066543, 0.881243724375963, 0.447462294017896, 0.212581163970754, 0.400236020097509, 0.913105392362922, 0.37836887454614, 0.464947072090581, 0.508277304237708, 0.0389843946322799, 0.00695240520872176, 0.775978752877563, 0.663806669181213, 0.435955132357776, 0.757332687266171, 0.766495012678206, 0.466838962864131, 0.0779885104857385, 0.65798934805207, 0.805620445637032, 0.885053988778964, 0.541377080604434, 0.0146494437940419, 0.905422102659941, 0.18866417510435, 0.0329374414868653, 0.750549020245671, 0.920892273308709, 0.407689163228497, 0.123866542009637, 0.25449550151825, 0.14396451972425, 0.780082255834714, 0.450569022679701, 0.558917294256389, 0.380576349329203, 0.605246999301016, 0.257436054525897, 0.208999912254512, 0.20464180316776, 0.792604504851624, 0.758399775018916, 0.225963254459202, 0.173042576061562, 0.0474998527206481, 0.499274828238413, 0.743603416951373, 0.416158657521009, 0.368043110705912, 0.0488379255402833, 0.134087378624827, 0.69313666713424, 0.863681641407311, 0.844314459245652, 0.125234917737544, 0.516750562470406, 0.46231782855466, 0.234701584093273, 0.400866707554087, 0.020523834740743, 0.947600563289598, 0.0889687987510115, 0.723583500832319, 0.355267892358825, 0.0895308454055339, 0.675794600509107, 0.539172405842692, 0.0638561288360506, 0.693931929534301, 0.609461208805442, 0.191566505236551, 0.989195969188586, 0.342337063048035, 0.371784995542839), actual = c(0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0), nmb = c(TP = -3, TN = 0, FP = -1, FN = -4)): could not find function "fx_roc01"